% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{abanyg}
\alias{abanyg}
\title{Abortion if woman wants for any reason (grid on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
abanyg
}
\details{
Question (Pleae indicate whether or not you think it should be possible for a pregnant woman to obtain a legal abortion . . .) If the woman wants it for any reason?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7624/vshow}{GSS Data Explorer page for \code{abanyg}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2714 \tab 779 \tab 521 \tab 4 \tab 1 \tab 13 \tab 4032 \cr
   2022 \tab 2380 \tab 704 \tab 423 \tab 29 \tab 5 \tab 3 \tab 3544 \cr
   Total \tab 5094 \tab 1483 \tab 944 \tab 33 \tab 6 \tab 16 \tab 7576 \cr
}
}

\keyword{variable}
