% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{abcare}
\alias{abcare}
\title{How concerned is r about abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abcare
}
\details{
Question How concerned are you personally about the abortion issue?  Are you very concerned, somewhat concerned, not very concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very concerned
\item \verb{[2]} somewhat concerned
\item \verb{[3]} not very concerned
\item \verb{[4]} not concerned at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/618/vshow}{GSS Data Explorer page for \code{abcare}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very concerned \tab somewhat concerned \tab not very concerned \tab not concerned at all \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 155 \tab 209 \tab 77 \tab 35 \tab 5 \tab 3 \tab 1473 \cr
   Total \tab 989 \tab 155 \tab 209 \tab 77 \tab 35 \tab 5 \tab 3 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{abfirm}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsoon}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{sexeduc}},
\code{\link{teenpill}}
}
\concept{Core}
\keyword{variable}
