% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_68.R
\name{abcon1}
\alias{abcon1}
\title{1st argument against abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abcon1
}
\details{
Question And, as far as you've heard, what are the main arguments against abortion?  PROBE FOR CLARITY, AND PROBE FOR ADDITIONAL REASONS.
A. First reason
}
\section{Values}{

\itemize{
\item \verb{[96]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/612/vshow}{GSS Data Explorer page for \code{abcon1}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 35 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 60 \tab none \tab no answer \tab Total \cr
   1982 \tab - \tab 438 \tab 288 \tab 45 \tab 9 \tab 110 \tab 24 \tab 233 \tab 5 \tab 19 \tab 102 \tab 31 \tab 15 \tab 12 \tab 4 \tab 39 \tab 4 \tab 12 \tab 115 \tab 36 \tab 21 \tab 4 \tab 5 \tab 35 \tab 251 \tab 3 \tab 1860 \cr
   Total \tab 0 \tab 438 \tab 288 \tab 45 \tab 9 \tab 110 \tab 24 \tab 233 \tab 5 \tab 19 \tab 102 \tab 31 \tab 15 \tab 12 \tab 4 \tab 39 \tab 4 \tab 12 \tab 115 \tab 36 \tab 21 \tab 4 \tab 5 \tab 35 \tab 251 \tab 3 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{aborct}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}}
}
\concept{Core}
\keyword{variable}
