% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_68.R
\name{abcon2}
\alias{abcon2}
\title{2nd argument against abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abcon2
}
\details{
Question And, as far as you've heard, what are the main arguments against abortion?  PROBE FOR CLARITY, AND PROBE FOR ADDITIONAL REASONS.
B. Second reason
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/613/vshow}{GSS Data Explorer page for \code{abcon2}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 35 \tab 4 \tab 41 \tab 42 \tab 43 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 60 \tab no answer \tab Total \cr
   1982 \tab 958 \tab 156 \tab 110 \tab 11 \tab 22 \tab 75 \tab 16 \tab 107 \tab 4 \tab 19 \tab 89 \tab 54 \tab 12 \tab 13 \tab 31 \tab 9 \tab 7 \tab 68 \tab 62 \tab 19 \tab 2 \tab 1 \tab 12 \tab 3 \tab 1860 \cr
   Total \tab 958 \tab 156 \tab 110 \tab 11 \tab 22 \tab 75 \tab 16 \tab 107 \tab 4 \tab 19 \tab 89 \tab 54 \tab 12 \tab 13 \tab 31 \tab 9 \tab 7 \tab 68 \tab 62 \tab 19 \tab 2 \tab 1 \tab 12 \tab 3 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcon1}},
\code{\link{abcon3}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{aborct}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}}
}
\concept{Core}
\keyword{variable}
