% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_626.R
\name{abdefectg}
\alias{abdefectg}
\title{Strong chance of serious defect (grid on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
abdefectg
}
\details{
Question (Pleae indicate whether or not you think it should be possible for a pregnant woman to obtain a legal abortion . . .) If there is a strong chance of serious defect in the baby?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7466/vshow}{GSS Data Explorer page for \code{abdefectg}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab don't know \tab no answer \tab Total \cr
   2021 \tab 2822 \tab 932 \tab 244 \tab 34 \tab - \tab - \tab 4032 \cr
   2022 \tab 2380 \tab 934 \tab 183 \tab 8 \tab 31 \tab 8 \tab 3544 \cr
   Total \tab 5202 \tab 1866 \tab 427 \tab 42 \tab 31 \tab 8 \tab 7576 \cr
}
}

\keyword{variable}
