% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{abfelegl}
\alias{abfelegl}
\title{Women only: women should be able to have legal abortions}
\source{
General Social Survey https://gss.norc.org
}
\description{
abfelegl
}
\details{
Question Leaving aside what you think of abortion for yourself, do you think a woman should continue to be able to have an abortion legally or not, or would you say it depends?
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not
\item \code{3} it depends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7172/vshow}{GSS Data Explorer page for \code{abfelegl}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab should \tab should not \tab it depends \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1052 \tab 481 \tab 193 \tab 581 \tab 9 \tab 32 \tab 2348 \cr
   Total \tab 1052 \tab 481 \tab 193 \tab 581 \tab 9 \tab 32 \tab 2348 \cr
}
}

\keyword{variable}
