% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{abfirm}
\alias{abfirm}
\title{How firm is r's opinion on abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abfirm
}
\details{
Question C. How firm are you about your opinion on abortion--would you say you are very likely to change your opinion, somewhat likely to change, somewhat unlikely to change or very unlikely to change?
}
\section{Values}{

\itemize{
\item \verb{[1]} change very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} change very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/617/vshow}{GSS Data Explorer page for \code{abfirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab change very likely \tab somewhat likely \tab somewhat unlikely \tab change very unlikely \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 48 \tab 215 \tab 413 \tab 1108 \tab 63 \tab 13 \tab 1860 \cr
   1984 \tab - \tab 44 \tab 195 \tab 335 \tab 869 \tab 23 \tab 7 \tab 1473 \cr
   Total \tab 0 \tab 92 \tab 410 \tab 748 \tab 1977 \tab 86 \tab 20 \tab 3333 \cr
}
}

\seealso{
Other Core: 
\code{\link{abcare}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsoon}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{sexeduc}},
\code{\link{teenpill}}
}
\concept{Core}
\keyword{variable}
