% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_661.R
\name{abgender}
\alias{abgender}
\title{Aboriton legal for gender selection}
\source{
General Social Survey https://gss.norc.org
}
\description{
abgender
}
\details{
Question (Please tell us whether or not you think it should be possible for a pregnant woman to obtain a legal abortion if...) ...she is pregnant with a boy but wants a girl, or is pregnant with a girl but wants a boy.

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7946/vshow}{GSS Data Explorer page for \code{abgender}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 182 \tab 1047 \tab 6 \tab 3544 \cr
   Total \tab 2309 \tab 182 \tab 1047 \tab 6 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{brv16bronum}},
\code{\link{brv16chinum}},
\code{\link{brv16cuznum}},
\code{\link{brv16frndnum}},
\code{\link{brv16othnum}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{gestate}}
}
\concept{GSS Next}
\keyword{variable}
