% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_68.R
\name{abinfo}
\alias{abinfo}
\title{How much info does r have on abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abinfo
}
\details{
Question B. How much information do you have about the abortion issue? Do you have all the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \verb{[1]} all info r needs
\item \verb{[2]} most info
\item \verb{[3]} some info
\item \verb{[4]} very little info
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/616/vshow}{GSS Data Explorer page for \code{abinfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all info r needs \tab most info \tab some info \tab very little info \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 406 \tab 366 \tab 597 \tab 467 \tab 15 \tab 9 \tab 1860 \cr
   1984 \tab - \tab 303 \tab 446 \tab 467 \tab 246 \tab 6 \tab 5 \tab 1473 \cr
   Total \tab 0 \tab 709 \tab 812 \tab 1064 \tab 713 \tab 21 \tab 14 \tab 3333 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abimp}},
\code{\link{aborct}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}}
}
\concept{Core}
\keyword{variable}
