% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_316.R
\name{abinspay}
\alias{abinspay}
\title{Should people be able to use health insurance for abortion - ver y}
\source{
General Social Survey https://gss.norc.org
}
\description{
abinspay
}
\details{
Question People use their health insurance to help cover the cost of receiving health care. Do you think people should be able to use their health insurance to help cover the cost of receiving an abortion?
}
\section{Values}{

\itemize{
\item \code{1} people should be able to
\item \code{2} people should not be able to
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7183/vshow}{GSS Data Explorer page for \code{abinspay}}.
}

\keyword{variable}
