% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_316.R
\name{abmedgov2}
\alias{abmedgov2}
\title{Woman and doctor or govt should decide what info needed for childbirth - ver y}
\source{
General Social Survey https://gss.norc.org
}
\description{
abmedgov2
}
\details{
Question Which of these statements comes closer to your view about what information a woman should have when she makes a decision about childbirth?
}
\section{Values}{

\itemize{
\item \code{1} a woman and her medical professional should decide
\item \code{2} the government should decide
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7182/vshow}{GSS Data Explorer page for \code{abmedgov2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab a woman and her medical professional should decide \tab the government should decide \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1188 \tab 1054 \tab 57 \tab 19 \tab 30 \tab 2348 \cr
   Total \tab 1188 \tab 1054 \tab 57 \tab 19 \tab 30 \tab 2348 \cr
}
}

\keyword{variable}
