% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{abmelegl}
\alias{abmelegl}
\title{Men only: women should be able to have legal abortions}
\source{
General Social Survey https://gss.norc.org
}
\description{
abmelegl
}
\details{
Question Leaving aside what you think of abortion for those close to you, do you think a woman should continue to be able to have an abortion legally or not or would you say it depends?
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not
\item \verb{[3]} it depends
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7173/vshow}{GSS Data Explorer page for \code{abmelegl}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab should \tab should not \tab it depends \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1296 \tab 423 \tab 134 \tab 464 \tab 9 \tab 22 \tab 2348 \cr
   Total \tab 1296 \tab 423 \tab 134 \tab 464 \tab 9 \tab 22 \tab 2348 \cr
}
}

\keyword{variable}
