% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{abmoral}
\alias{abmoral}
\title{R has a moral opposition to abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abmoral
}
\details{
Question Leaving aside whether or not you think abortion should be legal, are you morally opposed to abortion or not, or would you say it depends?
}
\section{Values}{

\itemize{
\item \verb{[1]} morally opposed
\item \verb{[2]} not morally opposed
\item \verb{[3]} it depends
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7174/vshow}{GSS Data Explorer page for \code{abmoral}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab morally opposed \tab not morally opposed \tab it depends \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 660 \tab 632 \tab 995 \tab 12 \tab 49 \tab - \tab 2348 \cr
   2022 \tab 2309 \tab 255 \tab 562 \tab 417 \tab - \tab - \tab 1 \tab 3544 \cr
   Total \tab 2309 \tab 915 \tab 1194 \tab 1412 \tab 12 \tab 49 \tab 1 \tab 5892 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{abhelp1}}
}
\concept{GSS Next}
\keyword{variable}
