% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{abnomore}
\alias{abnomore}
\title{Married--wants no more children}
\source{
General Social Survey https://gss.norc.org
}
\description{
abnomore
}
\details{
Question Please tell me whether or not you think it should be possible for a pregnant woman to obtain a legal abortion if. . .READ EACH STATEMENT, AND CIRCLE ONE CODE FOR EACH.
B. If she is married and does not want any more children?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/- \tab full \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
   1998 \tab A/C \tab partial \cr
   2000 \tab A/C \tab partial \cr
   2002 \tab A/C \tab partial \cr
   2004 \tab A/C \tab partial \cr
   2006 \tab A/C \tab partial \cr
   2008 \tab A/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
   2012 \tab A/C \tab partial \cr
   2014 \tab A/C \tab partial \cr
   2016 \tab A/C \tab partial \cr
   2018 \tab A/C \tab partial \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/602/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab no \tab no answer \tab yes \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 80 \tab 921 \tab 5 \tab 607 \tab - \tab - \tab 1613 \cr
   1973 \tab 49 \tab 760 \tab 2 \tab 693 \tab - \tab - \tab 1504 \cr
   1974 \tab 73 \tab 749 \tab - \tab 662 \tab - \tab - \tab 1484 \cr
   1975 \tab 62 \tab 774 \tab 2 \tab 652 \tab - \tab - \tab 1490 \cr
   1976 \tab 46 \tab 779 \tab 6 \tab 668 \tab - \tab - \tab 1499 \cr
   1977 \tab 61 \tab 782 \tab 7 \tab 680 \tab - \tab - \tab 1530 \cr
   1978 \tab 46 \tab 885 \tab 3 \tab 598 \tab - \tab - \tab 1532 \cr
   1980 \tab 58 \tab 744 \tab 3 \tab 663 \tab - \tab - \tab 1468 \cr
   1982 \tab 93 \tab 955 \tab 4 \tab 808 \tab - \tab - \tab 1860 \cr
   1983 \tab 49 \tab 928 \tab 30 \tab 592 \tab - \tab - \tab 1599 \cr
   1984 \tab 43 \tab 814 \tab 10 \tab 606 \tab - \tab - \tab 1473 \cr
   1985 \tab 41 \tab 888 \tab 5 \tab 600 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab 1470 \cr
   1987 \tab 68 \tab 1028 \tab 9 \tab 714 \tab - \tab - \tab 1819 \cr
   1988 \tab 26 \tab 570 \tab 2 \tab 379 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 37 \tab 550 \tab 5 \tab 441 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 33 \tab 483 \tab 5 \tab 396 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 36 \tab 526 \tab 7 \tab 424 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 51 \tab 539 \tab 5 \tab 480 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 65 \tab 995 \tab 6 \tab 930 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 91 \tab 972 \tab 7 \tab 853 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 86 \tab 1033 \tab 5 \tab 758 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 85 \tab 1050 \tab 6 \tab 720 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 29 \tab 493 \tab 2 \tab 400 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 30 \tab 502 \tab 6 \tab 360 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 65 \tab 1107 \tab 13 \tab 818 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 40 \tab 716 \tab 17 \tab 579 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 36 \tab 636 \tab 22 \tab 587 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 44 \tab 669 \tab 18 \tab 577 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 50 \tab 862 \tab 22 \tab 780 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 43 \tab 964 \tab 24 \tab 859 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 38 \tab 736 \tab 22 \tab 778 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 10 \tab 583 \tab 3 \tab 841 \tab 2574 \tab 21 \tab 4032 \cr
   2022 \tab 38 \tab 438 \tab 10 \tab 684 \tab 2367 \tab 7 \tab 3544 \cr
   2024 \tab 61 \tab 646 \tab 18 \tab 876 \tab 1694 \tab 14 \tab 3309 \cr
   Total \tab 1763 \tab 26077 \tab 311 \tab 22063 \tab 25443 \tab 42 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}

Other Abortion: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}}

Other Split Ballots: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}
}
\concept{Abortion}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
