% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_267.R
\name{abort12}
\alias{abort12}
\title{Past 12 months r web to learn abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abort12
}
\details{
Question In the past 12 months, have you used the Web to find out about or discuss?
B. Abortion issues?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2429/vshow}{GSS Data Explorer page for \code{abort12}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 29 \tab 251 \tab 84 \tab 2817 \cr
   2002 \tab 2411 \tab 30 \tab 279 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 59 \tab 530 \tab 129 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{chat12}},
\code{\link{econ12}},
\code{\link{game12}},
\code{\link{health12}},
\code{\link{locate12}},
\code{\link{newjob12}},
\code{\link{people12}},
\code{\link{polinf12}},
\code{\link{travel12}}
}
\concept{Information Society}
\keyword{variable}
