% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{abpoor}
\alias{abpoor}
\title{Low income--cant afford more children}
\source{
General Social Survey https://gss.norc.org
}
\description{
abpoor
}
\details{
Question Please tell me whether or not you think it should be possible for a pregnant woman to obtain a legal abortion if. . .READ EACH STATEMENT, AND CIRCLE ONE CODE FOR EACH.
D. If the family has a very low income and cannot afford any more children?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/604/vshow}{GSS Data Explorer page for \code{abpoor}}.

Counts by year:\tabular{llllllll}{
   year \tab yes \tab no \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 735 \tab 772 \tab 98 \tab 8 \tab - \tab - \tab 1613 \cr
   1973 \tab 778 \tab 678 \tab 46 \tab 2 \tab - \tab - \tab 1504 \cr
   1974 \tab 776 \tab 641 \tab 65 \tab 2 \tab - \tab - \tab 1484 \cr
   1975 \tab 753 \tab 663 \tab 69 \tab 5 \tab - \tab - \tab 1490 \cr
   1976 \tab 761 \tab 673 \tab 57 \tab 8 \tab - \tab - \tab 1499 \cr
   1977 \tab 789 \tab 689 \tab 44 \tab 8 \tab - \tab - \tab 1530 \cr
   1978 \tab 696 \tab 773 \tab 59 \tab 4 \tab - \tab - \tab 1532 \cr
   1980 \tab 728 \tab 680 \tab 58 \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab 868 \tab 905 \tab 79 \tab 8 \tab - \tab - \tab 1860 \cr
   1983 \tab 658 \tab 849 \tab 61 \tab 31 \tab - \tab - \tab 1599 \cr
   1984 \tab 654 \tab 760 \tab 45 \tab 14 \tab - \tab - \tab 1473 \cr
   1985 \tab 650 \tab 838 \tab 42 \tab 4 \tab - \tab - \tab 1534 \cr
   1987 \tab 778 \tab 964 \tab 69 \tab 8 \tab - \tab - \tab 1819 \cr
   1988 \tab 395 \tab 546 \tab 34 \tab 2 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 473 \tab 519 \tab 36 \tab 5 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 416 \tab 449 \tab 47 \tab 5 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 458 \tab 487 \tab 40 \tab 8 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 508 \tab 511 \tab 51 \tab 5 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 971 \tab 954 \tab 68 \tab 3 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 853 \tab 977 \tab 86 \tab 7 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 789 \tab 994 \tab 95 \tab 4 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 749 \tab 1024 \tab 81 \tab 7 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 396 \tab 496 \tab 29 \tab 3 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 352 \tab 506 \tab 29 \tab 11 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 822 \tab 1106 \tab 63 \tab 12 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 556 \tab 741 \tab 41 \tab 14 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 560 \tab 671 \tab 34 \tab 16 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 549 \tab 691 \tab 50 \tab 18 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 753 \tab 894 \tab 47 \tab 20 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 816 \tab 1012 \tab 37 \tab 25 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 750 \tab 770 \tab 36 \tab 18 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 825 \tab 600 \tab 13 \tab 3 \tab 2574 \tab 17 \tab 4032 \cr
   2022 \tab 680 \tab 446 \tab 32 \tab 9 \tab 2367 \tab 10 \tab 3544 \cr
   Total \tab 22295 \tab 24279 \tab 1741 \tab 299 \tab 22279 \tab 27 \tab 70920 \cr
}
}

\seealso{
Other Core: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}
}
\concept{Core}
\keyword{variable}
