% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{abpoor1}
\alias{abpoor1}
\title{Law should allow abortion if family low income}
\source{
General Social Survey https://gss.norc.org
}
\description{
abpoor1
}
\details{
Question Do you think the law should or should not allow a pregnant woman to obtain a legal abortion. . .
B. If the family has a very low income and cannot afford any more children.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should allow it
\item \verb{[2]} probably should allow it
\item \verb{[3]} probably should not allow it
\item \verb{[4]} definitely should not allow it
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4634/vshow}{GSS Data Explorer page for \code{abpoor1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should allow it \tab probably should allow it \tab probably should not allow it \tab definitely should not allow it \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 479 \tab 198 \tab 164 \tab 378 \tab 96 \tab 44 \tab 1517 \cr
   Total \tab 158 \tab 479 \tab 198 \tab 164 \tab 378 \tab 96 \tab 44 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}
}
\concept{ISSP Religion}
\keyword{variable}
