% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_68.R
\name{abpro1}
\alias{abpro1}
\title{1st argument for abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abpro1
}
\details{
Question As far as you've heard what are the main arguments in favor of abortions? PROBE EACH RESPONSE FOR CLARITY. PROBE FOR ADDITIONAL REASONS.
A. First reason
}
\section{Values}{

\itemize{
\item \code{96} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/609/vshow}{GSS Data Explorer page for \code{abpro1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 11 \tab 12 \tab 13 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 30 \tab 31 \tab 32 \tab 35 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 60 \tab none \tab no answer \tab Total \cr
   1982 \tab - \tab 52 \tab 159 \tab 189 \tab 48 \tab 8 \tab 59 \tab 184 \tab 22 \tab 16 \tab 78 \tab 3 \tab 24 \tab 12 \tab 205 \tab 3 \tab 195 \tab 8 \tab 41 \tab 77 \tab 17 \tab 12 \tab 8 \tab 12 \tab 40 \tab 5 \tab 15 \tab 366 \tab 2 \tab 1860 \cr
   Total \tab 0 \tab 52 \tab 159 \tab 189 \tab 48 \tab 8 \tab 59 \tab 184 \tab 22 \tab 16 \tab 78 \tab 3 \tab 24 \tab 12 \tab 205 \tab 3 \tab 195 \tab 8 \tab 41 \tab 77 \tab 17 \tab 12 \tab 8 \tab 12 \tab 40 \tab 5 \tab 15 \tab 366 \tab 2 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{aborct}},
\code{\link{abpro2}},
\code{\link{abpro3}}
}
\concept{Core}
\keyword{variable}
