% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{absent}
\alias{absent}
\title{How many days absent from work past 6 month}
\source{
General Social Survey https://gss.norc.org
}
\description{
absent
}
\details{
Question About how many days have you been absent from work in the last 6 months (not counting vacation)?
}
\section{Values}{

\itemize{
\item \verb{[1]} more thn 20 days
\item \verb{[2]} 11 to 20 days
\item \verb{[3]} 6 to 10 days
\item \verb{[4]} 1 to 5 days
\item \verb{[5]} none
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4590/vshow}{GSS Data Explorer page for \code{absent}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab more thn 20 days \tab 11 to 20 days \tab 6 to 10 days \tab 1 to 5 days \tab none \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 39 \tab 20 \tab 39 \tab 403 \tab 306 \tab 21 \tab 37 \tab 2832 \cr
   Total \tab 1967 \tab 39 \tab 20 \tab 39 \tab 403 \tab 306 \tab 21 \tab 37 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
