% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{absingle}
\alias{absingle}
\title{Not married}
\source{
General Social Survey https://gss.norc.org
}
\description{
absingle
}
\details{
Question Please tell me whether or not you think it should be possible for a pregnant woman to obtain a legal abortion if. . .READ EACH STATEMENT, AND CIRCLE ONE CODE FOR EACH.
F. If she is not married and does not want to marry the man?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/606/vshow}{GSS Data Explorer page for \code{absingle}}.

Counts by year:\tabular{llllllll}{
   year \tab yes \tab no \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 654 \tab 848 \tab 103 \tab 8 \tab - \tab - \tab 1613 \cr
   1973 \tab 712 \tab 738 \tab 49 \tab 5 \tab - \tab - \tab 1504 \cr
   1974 \tab 711 \tab 709 \tab 64 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 682 \tab 732 \tab 71 \tab 5 \tab - \tab - \tab 1490 \cr
   1976 \tab 723 \tab 715 \tab 56 \tab 5 \tab - \tab - \tab 1499 \cr
   1977 \tab 726 \tab 733 \tab 64 \tab 7 \tab - \tab - \tab 1530 \cr
   1978 \tab 606 \tab 867 \tab 55 \tab 4 \tab - \tab - \tab 1532 \cr
   1980 \tab 680 \tab 724 \tab 62 \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab 803 \tab 961 \tab 91 \tab 5 \tab - \tab - \tab 1860 \cr
   1983 \tab 588 \tab 905 \tab 73 \tab 33 \tab - \tab - \tab 1599 \cr
   1984 \tab 628 \tab 792 \tab 42 \tab 11 \tab - \tab - \tab 1473 \cr
   1985 \tab 612 \tab 874 \tab 43 \tab 5 \tab - \tab - \tab 1534 \cr
   1987 \tab 700 \tab 1045 \tab 65 \tab 9 \tab - \tab - \tab 1819 \cr
   1988 \tab 367 \tab 565 \tab 41 \tab 4 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 447 \tab 538 \tab 43 \tab 5 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 395 \tab 477 \tab 41 \tab 4 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 426 \tab 524 \tab 36 \tab 7 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 488 \tab 526 \tab 56 \tab 5 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 923 \tab 1015 \tab 52 \tab 6 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 822 \tab 1010 \tab 80 \tab 11 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 757 \tab 1033 \tab 87 \tab 5 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 691 \tab 1076 \tab 88 \tab 6 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 376 \tab 519 \tab 27 \tab 2 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 354 \tab 512 \tab 20 \tab 12 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 777 \tab 1159 \tab 57 \tab 10 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 536 \tab 765 \tab 38 \tab 13 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 524 \tab 704 \tab 36 \tab 17 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 533 \tab 713 \tab 45 \tab 17 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 713 \tab 938 \tab 44 \tab 19 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 782 \tab 1051 \tab 30 \tab 27 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 700 \tab 816 \tab 38 \tab 20 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 813 \tab 612 \tab 11 \tab 1 \tab 2574 \tab 21 \tab 4032 \cr
   2022 \tab 666 \tab 467 \tab 29 \tab 10 \tab 2367 \tab 5 \tab 3544 \cr
   Total \tab 20915 \tab 25663 \tab 1737 \tab 300 \tab 22279 \tab 26 \tab 70920 \cr
}
}

\seealso{
Other Core: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}
}
\concept{Core}
\keyword{variable}
