% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{absingleg}
\alias{absingleg}
\title{Not married (grid on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
absingleg
}
\details{
Question (Pleae indicate whether or not you think it should be possible for a pregnant woman to obtain a legal abortion . . .) If she is not married and does not want to marry the man?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7471/vshow}{GSS Data Explorer page for \code{absingleg}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab don't know \tab no answer \tab Total \cr
   2021 \tab 2822 \tab 624 \tab 534 \tab 52 \tab - \tab - \tab 4032 \cr
   2022 \tab 2380 \tab 667 \tab 445 \tab 14 \tab 33 \tab 5 \tab 3544 \cr
   Total \tab 5202 \tab 1291 \tab 979 \tab 66 \tab 33 \tab 5 \tab 7576 \cr
}
}

\keyword{variable}
