% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{abspno}
\alias{abspno}
\title{Abortion if husband opposes}
\source{
General Social Survey https://gss.norc.org
}
\description{
abspno
}
\details{
Question If a woman has decided to have an abortion but her husband is against it, do you think she should go ahead and have  it, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/698/vshow}{GSS Data Explorer page for \code{abspno}}.
}

\seealso{
Other Core: 
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{comprend}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{form}},
\code{\link{phone}},
\code{\link{radiohrs}}

Other Abortion: 
\code{\link{abhave1}},
\code{\link{abhave2}}
}
\concept{Abortion}
\concept{Core}
\keyword{variable}
