% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{abstate1}
\alias{abstate1}
\title{Difficulty of obtaining abortion in r's state}
\source{
General Social Survey https://gss.norc.org
}
\description{
abstate1
}
\details{
Question Here in (CURRENT STATE), how easy or hard do you think it is for a woman to get an abortion?
}
\section{Values}{

\itemize{
\item \verb{[1]} very easy
\item \verb{[2]} easy
\item \verb{[3]} neither easy nor hard
\item \verb{[4]} hard
\item \verb{[5]} very hard
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7175/vshow}{GSS Data Explorer page for \code{abstate1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very easy \tab easy \tab neither easy nor hard \tab hard \tab very hard \tab don't know \tab no answer \tab Total \cr
   2018 \tab - \tab 367 \tab 671 \tab 518 \tab 217 \tab 83 \tab 442 \tab 50 \tab 2348 \cr
   Total \tab 0 \tab 367 \tab 671 \tab 518 \tab 217 \tab 83 \tab 442 \tab 50 \tab 2348 \cr
}
}

\keyword{variable}
