% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{abstate2}
\alias{abstate2}
\title{Change abortion laws in r's state to make it easier or harder to get an abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abstate2
}
\details{
Question Here in (CURRENT STATE), do you think that laws should be changed to make it easier for a woman to get an abortion, be changed to make it harder for a woman to get an abortion, or should the laws stay as they are now?
}
\section{Values}{

\itemize{
\item \verb{[1]} make it easier
\item \verb{[2]} make it harder
\item \verb{[3]} stay the same as now
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7176/vshow}{GSS Data Explorer page for \code{abstate2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab make it easier \tab make it harder \tab stay the same as now \tab don't know \tab no answer \tab Total \cr
   2018 \tab - \tab 403 \tab 610 \tab 1046 \tab 235 \tab 54 \tab 2348 \cr
   Total \tab 0 \tab 403 \tab 610 \tab 1046 \tab 235 \tab 54 \tab 2348 \cr
}
}

\keyword{variable}
