% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{accntsci}
\alias{accntsci}
\title{How scientific: accounting}
\source{
General Social Survey https://gss.norc.org
}
\description{
accntsci
}
\details{
Question Please look at Card B20. How scientific are each of the following fields? If you have not heard of a particular field, just say you haven’t heard of it
D. Accounting. Is accounting very scientific, pretty scientific, not too scientific, or not scientific at all?
}
\section{Values}{

\itemize{
\item \code{1} very scientific
\item \code{2} pretty scientific
\item \code{3} not too scientific
\item \code{4} not scientific at all
\item \code{5} haven't heard of it (volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3409/vshow}{GSS Data Explorer page for \code{accntsci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab haven't heard of it (volunteered) \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 268 \tab 377 \tab 555 \tab 585 \tab 12 \tab 63 \tab 4 \tab 4510 \cr
   2012 \tab 972 \tab 100 \tab 199 \tab 348 \tab 307 \tab 4 \tab 37 \tab 7 \tab 1974 \cr
   Total \tab 3618 \tab 368 \tab 576 \tab 903 \tab 892 \tab 16 \tab 100 \tab 11 \tab 6484 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{biosci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{medsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
