% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{accptoth}
\alias{accptoth}
\title{R accept others even when they do things wrong}
\source{
General Social Survey https://gss.norc.org
}
\description{
accptoth
}
\details{
Question The following are things that you may experience in your daily life. Please tell me how often these occur.
B. I accept others even when they do things I think are wrong.
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never or almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2889/vshow}{GSS Data Explorer page for \code{accptoth}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never or almost never \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2002 \tab 1393 \tab 119 \tab 207 \tab 437 \tab 310 \tab 206 \tab 66 \tab 20 \tab 7 \tab - \tab 2765 \cr
   2004 \tab 1472 \tab 144 \tab 365 \tab 451 \tab 212 \tab 124 \tab 20 \tab 10 \tab 14 \tab - \tab 2812 \cr
   2012 \tab 666 \tab 136 \tab 281 \tab 389 \tab 223 \tab 189 \tab 76 \tab 8 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 145 \tab 255 \tab 373 \tab 222 \tab 186 \tab 79 \tab 5 \tab 9 \tab - \tab 2538 \cr
   Total \tab 4795 \tab 544 \tab 1108 \tab 1650 \tab 967 \tab 705 \tab 241 \tab 43 \tab 36 \tab 0 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{careself}},
\code{\link{carried}},
\code{\link{cutahead}},
\code{\link{directns}},
\code{\link{empathy1}},
\code{\link{empathy2}},
\code{\link{empathy3}},
\code{\link{empathy4}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{firstyou}},
\code{\link{givblood}},
\code{\link{givchrty}},
\code{\link{givhmlss}},
\code{\link{givseat}},
\code{\link{helpaway}},
\code{\link{helpfrds}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{retchnge}},
\code{\link{selffrst}},
\code{\link{selfless}},
\code{\link{usefrds}},
\code{\link{valgiven}},
\code{\link{volchrty}},
\code{\link{volmonth}}

Other Feelings: 
\code{\link{afailure}},
\code{\link{angrlast}},
\code{\link{angry}},
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{anxious}},
\code{\link{ashamed}},
\code{\link{atease}},
\code{\link{badbrks}},
\code{\link{beplesnt}},
\code{\link{big5a1}},
\code{\link{big5a2}},
\code{\link{big5b1}},
\code{\link{big5b2}},
\code{\link{big5c1}},
\code{\link{big5c2}},
\code{\link{big5d1}},
\code{\link{big5d2}},
\code{\link{big5e1}},
\code{\link{big5e2}},
\code{\link{calm}},
\code{\link{chnang1}},
\code{\link{chnang10}},
\code{\link{chnang11}},
\code{\link{chnang12}},
\code{\link{chnang13}},
\code{\link{chnang14}},
\code{\link{chnang15}},
\code{\link{chnang16}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{chnang6}},
\code{\link{chnang7}},
\code{\link{chnang8}},
\code{\link{chnang9}},
\code{\link{contentd}},
\code{\link{effort}},
\code{\link{embarrss}},
\code{\link{empathy1}},
\code{\link{empathy2}},
\code{\link{empathy3}},
\code{\link{empathy4}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{fidgety}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{hapfeel}},
\code{\link{hopeless}},
\code{\link{howangry}},
\code{\link{likeoth}},
\code{\link{litcntrl}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{madat1}},
\code{\link{madat10}},
\code{\link{madat11}},
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat19}},
\code{\link{madat2}},
\code{\link{madat20}},
\code{\link{madat21}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat5}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}},
\code{\link{moregood}},
\code{\link{mostluck}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{noemote}},
\code{\link{nogood}},
\code{\link{noplan}},
\code{\link{notcount}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{othresp}},
\code{\link{outraged}},
\code{\link{ovrjoyed}},
\code{\link{pessimst}},
\code{\link{proud}},
\code{\link{reactok}},
\code{\link{restless}},
\code{\link{sad}},
\code{\link{satself}},
\code{\link{selfless}},
\code{\link{selfresp}},
\code{\link{shakeblu}},
\code{\link{showangr}},
\code{\link{showfeel}},
\code{\link{slfrspct}},
\code{\link{thnkangr}},
\code{\link{whnangry}},
\code{\link{worried}},
\code{\link{wrthless}}
}
\concept{Altruism}
\concept{Feelings}
\keyword{variable}
