% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{acpttg}
\alias{acpttg}
\title{Accepting people}
\source{
General Social Survey https://gss.norc.org
}
\description{
acpttg
}
\details{
Question Do you think American society has gone too far, about far enough, or not far enough in accepting people (whose sex differs from the one they were assigned at birth)?
Question edits are in compliance with federal administration directives released in January 2025.
}
\section{Values}{

\itemize{
\item \verb{[1]} too far
\item \verb{[2]} about far enough
\item \verb{[3]} not far enough
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8217/vshow}{GSS Data Explorer page for \code{acpttg}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab too far \tab about far enough \tab not far enough \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 908 \tab 559 \tab 602 \tab 86 \tab 1106 \tab 15 \tab 33 \tab 3309 \cr
   Total \tab 0 \tab 908 \tab 559 \tab 602 \tab 86 \tab 1106 \tab 15 \tab 33 \tab 3309 \cr
}
}

\seealso{
Other Sex: 
\code{\link{chngdocs}}
}
\concept{Sex}
\keyword{variable}
