% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_393.R
\name{acqbrnda}
\alias{acqbrnda}
\title{Number of brenda r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqbrnda
}
\details{
Question Cases in Sample B (Ballots 4-6) should receive this variable. This variable is used to drive skips in a module. This  variable should have a value of 1-5 using the following proportions: 1=50pct, 2,3,4,5 = 12.5pct each.
D. How many are named Brenda?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3575/vshow}{GSS Data Explorer page for \code{acqbrnda}}.
}

\seealso{
Other Number Known: 
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqmark}},
\code{\link{acqrachl}},
\code{\link{acqshawn}},
\code{\link{numknown}}
}
\concept{Number Known}
\keyword{variable}
