% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_395.R
\name{acqcops}
\alias{acqcops}
\title{Number of police officer r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqcops
}
\details{
Question A. How many of the people you are acquainted with are police officers?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3595/vshow}{GSS Data Explorer page for \code{acqcops}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3828 \tab 266 \tab 133 \tab 199 \tab 37 \tab 39 \tab 1 \tab 7 \tab 4510 \cr
   Total \tab 3828 \tab 266 \tab 133 \tab 199 \tab 37 \tab 39 \tab 1 \tab 7 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqjans}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqnoatt}},
\code{\link{acqsocs}}
}
\concept{Number Known}
\keyword{variable}
