% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_402.R
\name{acqdems}
\alias{acqdems}
\title{Number of democrat r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqdems
}
\details{
Question Now thinking about all the people you are acquainted with:
B. How many of do you think are Democrats? Would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} almost all
\item \verb{[2]} most
\item \verb{[3]} about half
\item \verb{[4]} a few
\item \verb{[5]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3663/vshow}{GSS Data Explorer page for \code{acqdems}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost all \tab most \tab about half \tab a few \tab none \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3139 \tab 106 \tab 165 \tab 555 \tab 292 \tab 29 \tab 204 \tab 20 \tab 4510 \cr
   Total \tab 3139 \tab 106 \tab 165 \tab 555 \tab 292 \tab 29 \tab 204 \tab 20 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqattnd}},
\code{\link{acqmyrac}},
\code{\link{acqreps}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvalib}},
\code{\link{trtkaren}},
\code{\link{trtkevin}},
\code{\link{trtshawn}}
}
\concept{Number Known}
\keyword{variable}
