% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_397.R
\name{acqfmgay}
\alias{acqfmgay}
\title{Number of gay family member}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqfmgay
}
\details{
Question Next, we are going to ask questions about people in your family, including relatives and in-laws.
J. How many are gay men or women?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3611/vshow}{GSS Data Explorer page for \code{acqfmgay}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3996 \tab 341 \tab 86 \tab 58 \tab 3 \tab 3 \tab 14 \tab 9 \tab 4510 \cr
   Total \tab 3996 \tab 341 \tab 86 \tab 58 \tab 3 \tab 3 \tab 14 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgo}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmno}},
\code{\link{acqfmwht}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}}
}
\concept{Number Known}
\keyword{variable}
