% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_396.R
\name{acqfmhme}
\alias{acqfmhme}
\title{Number of family member owning second home}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqfmhme
}
\details{
Question Next, we are going to ask questions about people in your family, including relatives and in-laws.
D. How many own a second home?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3605/vshow}{GSS Data Explorer page for \code{acqfmhme}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3996 \tab 264 \tab 97 \tab 118 \tab 9 \tab 3 \tab 17 \tab 6 \tab 4510 \cr
   Total \tab 3996 \tab 264 \tab 97 \tab 118 \tab 9 \tab 3 \tab 17 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqchild}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqmils}}
}
\concept{Number Known}
\keyword{variable}
