% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_396.R
\name{acqfmmrk}
\alias{acqfmmrk}
\title{Number of family member named mark}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqfmmrk
}
\details{
Question Next, we are going to ask questions about people in your family, including relatives and in-laws.
A. How many people in your family, including relatives and in-laws, are named Mark?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3602/vshow}{GSS Data Explorer page for \code{acqfmmrk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3996 \tab 315 \tab 153 \tab 38 \tab 2 \tab 6 \tab 4510 \cr
   Total \tab 3996 \tab 315 \tab 153 \tab 38 \tab 2 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqchild}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmhme}},
\code{\link{acqfmlin}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqmils}}
}
\concept{Number Known}
\keyword{variable}
