% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_395.R
\name{acqjans}
\alias{acqjans}
\title{Number of janitor r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqjans
}
\details{
Question D. How many are janitors or building cleaners?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3598/vshow}{GSS Data Explorer page for \code{acqjans}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3828 \tab 301 \tab 140 \tab 185 \tab 27 \tab 18 \tab 4 \tab 7 \tab 4510 \cr
   Total \tab 3828 \tab 301 \tab 140 \tab 185 \tab 27 \tab 18 \tab 4 \tab 7 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqnoatt}},
\code{\link{acqsocs}}
}
\concept{Number Known}
\keyword{variable}
