% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_394.R
\name{acqlinda}
\alias{acqlinda}
\title{Number of linda r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqlinda
}
\details{
Question Cases in Sample B (Ballots 4-6) should receive this variable. This variable is used to drive skips in a module. This  variable should have a value of 1-5 using the following proportions: 1=50pct, 2,3,4,5 = 12.5pct each.
H. How many are named Linda?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3579/vshow}{GSS Data Explorer page for \code{acqlinda}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 171 \tab 206 \tab 254 \tab 30 \tab 8 \tab 6 \tab 14 \tab 4510 \cr
   Total \tab 3821 \tab 171 \tab 206 \tab 254 \tab 30 \tab 8 \tab 6 \tab 14 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqasian}},
\code{\link{acqblack}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjose}},
\code{\link{acqmaria}},
\code{\link{acqprisn}},
\code{\link{acqunemp}},
\code{\link{acqwhite}}
}
\concept{Number Known}
\keyword{variable}
