% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_393.R
\name{acqmark}
\alias{acqmark}
\title{Number of mark r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqmark
}
\details{
Question Cases in Sample B (Ballots 4-6) should receive this variable. This variable is used to drive skips in a module. This  variable should have a value of 1-5 using the following proportions: 1=50pct, 2,3,4,5 = 12.5pct each.
G. How many are named Mark?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3578/vshow}{GSS Data Explorer page for \code{acqmark}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 165 \tab 183 \tab 266 \tab 38 \tab 19 \tab 4 \tab 14 \tab 4510 \cr
   Total \tab 3821 \tab 165 \tab 183 \tab 266 \tab 38 \tab 19 \tab 4 \tab 14 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqbrnda}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqrachl}},
\code{\link{acqshawn}},
\code{\link{numknown}}
}
\concept{Number Known}
\keyword{variable}
