% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_402.R
\name{acqmyrac}
\alias{acqmyrac}
\title{Proportion of same race as r among all acquaintances}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqmyrac
}
\details{
Question Are the people that you are acquainted with…
}
\section{Values}{

\itemize{
\item \verb{[1]} almost all the same race as you
\item \verb{[2]} mostly the same race as you
\item \verb{[3]} about evenly divided
\item \verb{[4]} mostly a different race than you
\item \verb{[5]} almost all a different race than you
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3665/vshow}{GSS Data Explorer page for \code{acqmyrac}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost all the same race as you \tab mostly the same race as you \tab about evenly divided \tab mostly a different race than you \tab almost all a different race than you \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3139 \tab 447 \tab 530 \tab 270 \tab 57 \tab 18 \tab 29 \tab 20 \tab 4510 \cr
   Total \tab 3139 \tab 447 \tab 530 \tab 270 \tab 57 \tab 18 \tab 29 \tab 20 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqattnd}},
\code{\link{acqdems}},
\code{\link{acqreps}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvalib}},
\code{\link{trtkaren}},
\code{\link{trtkevin}},
\code{\link{trtshawn}}
}
\concept{Number Known}
\keyword{variable}
