% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_399.R
\name{acqnhcoh}
\alias{acqnhcoh}
\title{Number of cohabiting ppl in neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqnhcoh
}
\details{
Question Next, we are going to ask questions about people that you are acquainted with through your neighborhood.
O. How many women that you are acquainted with in your neighborhood are you pretty certain live in a romantic relationship with a man to whom they are not married?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3631/vshow}{GSS Data Explorer page for \code{acqnhcoh}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 4002 \tab 259 \tab 52 \tab 66 \tab 2 \tab 10 \tab 113 \tab 6 \tab 4510 \cr
   Total \tab 4002 \tab 259 \tab 52 \tab 66 \tab 2 \tab 10 \tab 113 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqnhcon}},
\code{\link{acqnhlib}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkhme}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}}
}
\concept{Number Known}
\keyword{variable}
