% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_397.R
\name{acqnhlin}
\alias{acqnhlin}
\title{Number of linda in neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqnhlin
}
\details{
Question Next, we are going to ask questions about people that you are acquainted with through your neighborhood.
B. How many of the people you are acquainted with through your neighborhood are named Linda?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3618/vshow}{GSS Data Explorer page for \code{acqnhlin}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 4002 \tab 380 \tab 80 \tab 29 \tab 2 \tab 11 \tab 6 \tab 4510 \cr
   Total \tab 4002 \tab 380 \tab 80 \tab 29 \tab 2 \tab 11 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmno}},
\code{\link{acqfmwht}},
\code{\link{acqnhmrk}}
}
\concept{Number Known}
\keyword{variable}
