% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_398.R
\name{acqnhwht}
\alias{acqnhwht}
\title{Number of white in neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqnhwht
}
\details{
Question Next, we are going to ask questions about people that you are acquainted with through your neighborhood.
I. How many are white?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3625/vshow}{GSS Data Explorer page for \code{acqnhwht}}.
}

\seealso{
Other Number Known: 
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}}

Other Social Networks: 
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}}
}
\concept{Number Known}
\concept{Social Networks}
\concept{Whites}
\keyword{variable}
