% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_395.R
\name{acqsocs}
\alias{acqsocs}
\title{Number of social worker r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqsocs
}
\details{
Question C. How many are social workers?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3597/vshow}{GSS Data Explorer page for \code{acqsocs}}.
}

\seealso{
Other Number Known: 
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqjans}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqnoatt}}
}
\concept{Number Known}
\keyword{variable}
