% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_402.R
\name{acqvacoh}
\alias{acqvacoh}
\title{Number of cohabiting ppl in voluntary association}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqvacoh
}
\details{
Question Next, we are going to ask about people that you are acquainted with through schools, clubs, associations, or places of worship.
O. How many women that you are acquainted with through schools, clubs, associations, or places of worship are you pretty certain live in a romantic relationship with a man to whom they are not married?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3661/vshow}{GSS Data Explorer page for \code{acqvacoh}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3987 \tab 211 \tab 46 \tab 92 \tab 21 \tab 36 \tab 108 \tab 9 \tab 4510 \cr
   Total \tab 3987 \tab 211 \tab 46 \tab 92 \tab 21 \tab 36 \tab 108 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqattnd}},
\code{\link{acqdems}},
\code{\link{acqmyrac}},
\code{\link{acqreps}},
\code{\link{acqvacon}},
\code{\link{acqvalib}},
\code{\link{trtkaren}},
\code{\link{trtkevin}},
\code{\link{trtshawn}}
}
\concept{Number Known}
\keyword{variable}
