% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_401.R
\name{acqvago}
\alias{acqvago}
\title{Number of regular church attendee in voluntary association}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqvago
}
\details{
Question Next, we are going to ask about people that you are acquainted with through schools, clubs, associations, or places of worship.
K. How many are you pretty certain attend religious services on a regular basis?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3657/vshow}{GSS Data Explorer page for \code{acqvago}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3987 \tab 80 \tab 27 \tab 74 \tab 43 \tab 181 \tab 110 \tab 8 \tab 4510 \cr
   Total \tab 3987 \tab 80 \tab 27 \tab 74 \tab 43 \tab 181 \tab 110 \tab 8 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvagay}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}}
}
\concept{Number Known}
\keyword{variable}
