% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_401.R
\name{acqvahsp}
\alias{acqvahsp}
\title{Number of hispanic in voluntary association}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqvahsp
}
\details{
Question Next, we are going to ask about people that you are acquainted with through schools, clubs, associations, or places of worship.
H. How many are Hispanic?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3654/vshow}{GSS Data Explorer page for \code{acqvahsp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3987 \tab 242 \tab 31 \tab 124 \tab 35 \tab 56 \tab 27 \tab 8 \tab 4510 \cr
   Total \tab 3987 \tab 242 \tab 31 \tab 124 \tab 35 \tab 56 \tab 27 \tab 8 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}}
}
\concept{Number Known}
\keyword{variable}
