% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_400.R
\name{acqwkgay}
\alias{acqwkgay}
\title{Number of gay in current or previous work}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqwkgay
}
\details{
Question Next, we are going to ask about people that you are acquainted with through work.
J. How many are you pretty certain are gay men or women?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{6} i dont have a job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3641/vshow}{GSS Data Explorer page for \code{acqwkgay}}.
}

\seealso{
Other Number Known: 
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgo}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwkno}},
\code{\link{acqwkwht}}

Other Social Networks: 
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgo}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwkno}},
\code{\link{acqwkwht}}
}
\concept{Gays And Lesbians}
\concept{Number Known}
\concept{Social Networks}
\keyword{variable}
