% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_400.R
\name{acqwkwht}
\alias{acqwkwht}
\title{Number of white in current or previous work}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqwkwht
}
\details{
Question Next, we are going to ask about people that you are acquainted with through work.
I. How many are white?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[6]} i dont have a job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3640/vshow}{GSS Data Explorer page for \code{acqwkwht}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab I don't have a job \tab no answer \tab Total \cr
   2006 \tab 4009 \tab 34 \tab 14 \tab 41 \tab 47 \tab 330 \tab 16 \tab 5 \tab 14 \tab 4510 \cr
   Total \tab 4009 \tab 34 \tab 14 \tab 41 \tab 47 \tab 330 \tab 16 \tab 5 \tab 14 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwkno}}
}
\concept{Number Known}
\keyword{variable}
