% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{actassoc}
\alias{actassoc}
\title{How important to be active on soc or pol association}
\source{
General Social Survey https://gss.norc.org
}
\description{
actassoc
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
E. To be active in social or political associations
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4911/vshow}{GSS Data Explorer page for \code{actassoc}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 93 \tab 71 \tab 146 \tab 340 \tab 373 \tab 196 \tab 234 \tab 16 \tab 3 \tab 2812 \cr
   2014 \tab 1274 \tab 84 \tab 48 \tab 118 \tab 272 \tab 317 \tab 179 \tab 195 \tab 48 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 177 \tab 119 \tab 264 \tab 612 \tab 690 \tab 375 \tab 429 \tab 64 \tab 6 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{relmeet}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
