% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_109.R
\name{acthobby}
\alias{acthobby}
\title{R active in hobby club}
\source{
General Social Survey https://gss.norc.org
}
\description{
acthobby
}
\details{
Question Have you ever done any active work for (EACH "YES" IN MEMFRAT-MEMOTHER)?  I mean been a leader, helped organize meetings, been an officer, or given time or money?
I. Hobby or garden clubs
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1004/vshow}{GSS Data Explorer page for \code{acthobby}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab 1650 \tab 85 \tab 51 \tab 33 \tab 1819 \cr
   Total \tab 1650 \tab 85 \tab 51 \tab 33 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actfarm}},
\code{\link{actgreek}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actyouth}}
}
\concept{Socio-political participation}
\keyword{variable}
