% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{activnat}
\alias{activnat}
\title{Outside leisure activity in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
activnat
}
\details{
Question In the last twelve months how often, if at all, have you engaged in any leisure activities outside in nature, such as hiking, bird watching, swimming, skiing, other outdoor activities, or just relaxing?
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} several times a week
\item \code{3} several times a month
\item \code{4} several times a year
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7496/vshow}{GSS Data Explorer page for \code{activnat}}.
}

\keyword{variable}
