% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_109.R
\name{actnat}
\alias{actnat}
\title{R active in nationality group}
\source{
General Social Survey https://gss.norc.org
}
\description{
actnat
}
\details{
Question Have you ever done any active work for (EACH "YES" IN MEMFRAT-MEMOTHER)?  I mean been a leader, helped organize meetings, been an officer, or given time or money?
K. Nationality groups
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1006/vshow}{GSS Data Explorer page for \code{actnat}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab 1751 \tab 30 \tab 12 \tab 26 \tab 1819 \cr
   Total \tab 1751 \tab 30 \tab 12 \tab 26 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actfarm}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actyouth}}
}
\concept{Socio-political participation}
\keyword{variable}
