% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_109.R
\name{actschl}
\alias{actschl}
\title{R active in school service}
\source{
General Social Survey https://gss.norc.org
}
\description{
actschl
}
\details{
Question Have you ever done any active work for (EACH "YES" IN MEMFRAT-MEMOTHER)?  I mean been a leader, helped organize meetings, been an officer, or given time or money?
H. School service groups
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1003/vshow}{GSS Data Explorer page for \code{actschl}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{actfarm}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actprof}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actyouth}}
}
\concept{Socio-political participation}
\keyword{variable}
