% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_329.R
\name{adcou}
\alias{adcou}
\title{Would r go to counselor if child was r's}
\source{
General Social Survey https://gss.norc.org
}
\description{
adcou
}
\details{
Question If (John/Mary) was your child, would you...
Go to a therapist or counselor, like a psychologist, social worker, or other mental health professional for help?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2943/vshow}{GSS Data Explorer page for \code{adcou}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 941 \tab 415 \tab 26 \tab 11 \tab 2765 \cr
   Total \tab 1372 \tab 941 \tab 415 \tab 26 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldfrnd}},
\code{\link{chldsch}},
\code{\link{spendeve}}
}
\concept{Mental Health II}
\keyword{variable}
