% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_329.R
\name{addoc}
\alias{addoc}
\title{Would go to a medical doctor if he was r's child}
\source{
General Social Survey https://gss.norc.org
}
\description{
addoc
}
\details{
Question If (John/Mary) was your child, would you...
Go to a general medical doctor for help?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2941/vshow}{GSS Data Explorer page for \code{addoc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2002 \tab 1372 \tab 1091 \tab 268 \tab 23 \tab 11 \tab - \tab - \tab 2765 \cr
   2024 \tab 1661 \tab 1224 \tab 379 \tab 15 \tab 5 \tab - \tab 25 \tab 3309 \cr
   Total \tab 3033 \tab 2315 \tab 647 \tab 38 \tab 16 \tab 0 \tab 25 \tab 6074 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldfrnd}},
\code{\link{chldsch}},
\code{\link{spendeve}}

Other Children: 
\code{\link{adcou}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldfrnd}},
\code{\link{chldsch}},
\code{\link{spendeve}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
