% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_329.R
\name{addocmed}
\alias{addocmed}
\title{Would have child be on med by docs advice}
\source{
General Social Survey https://gss.norc.org
}
\description{
addocmed
}
\details{
Question If (John/Mary) was your child, would you...
Go to a general medical doctor for help?
A. If the doctor said that your child should be on medication, how likely would you be to do that? Very likely, somewhat likely, not very likely, or not at all likely?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2942/vshow}{GSS Data Explorer page for \code{addocmed}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{addoc}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldfrnd}},
\code{\link{chldsch}},
\code{\link{spendeve}}
}
\concept{Mental Health II}
\keyword{variable}
