% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_329.R
\name{adfammed}
\alias{adfammed}
\title{Would r have child be on medication by fam advice}
\source{
General Social Survey https://gss.norc.org
}
\description{
adfammed
}
\details{
Question If (John/Mary) was your child, would you...
Get advice from family and friends about (John/Mary)?
A. If they said your child should be on medication, how likely would you be to do that?  Very likely, somewhat likely, not very likely, or not at all likely?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2938/vshow}{GSS Data Explorer page for \code{adfammed}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab Total \cr
   2002 \tab 1844 \tab 273 \tab 279 \tab 216 \tab 145 \tab 8 \tab 2765 \cr
   Total \tab 1844 \tab 273 \tab 279 \tab 216 \tab 145 \tab 8 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldfrnd}},
\code{\link{chldsch}},
\code{\link{spendeve}}
}
\concept{Mental Health II}
\keyword{variable}
