% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{adforjob}
\alias{adforjob}
\title{Advertised for a job in newspaper's}
\source{
General Social Survey https://gss.norc.org
}
\description{
adforjob
}
\details{
Question Thinking about the last 12 months, have you done any of the following in order to find a job?
D. Advertised for a job in newspapers or journals?
}
\section{Values}{

\itemize{
\item \code{1} no
\item \code{2} yes, once or twice
\item \code{3} yes, more than twice
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4602/vshow}{GSS Data Explorer page for \code{adforjob}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab yes, once or twice \tab yes, more than twice \tab no answer \tab Total \cr
   1998 \tab 2460 \tab 278 \tab 14 \tab 8 \tab 72 \tab 2832 \cr
   2006 \tab 4008 \tab 478 \tab 11 \tab 9 \tab 4 \tab 4510 \cr
   2016 \tab 2533 \tab 268 \tab 22 \tab 21 \tab 23 \tab 2867 \cr
   Total \tab 9001 \tab 1024 \tab 47 \tab 38 \tab 99 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{econsup}},
\code{\link{rimpskls}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}
}
\concept{Work Orientation}
\keyword{variable}
