% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{adhdknow}
\alias{adhdknow}
\title{What does r know about adhd}
\source{
General Social Survey https://gss.norc.org
}
\description{
adhdknow
}
\details{
Question Please tell me what you know about Attention Deficit/Hyperactivity Disorder or ADHD. PROBE: Anything else?
}
\section{Values}{

\itemize{
\item \code{1} never heard of it
\item \code{2} not sure/dont know
\item \code{3} no answer
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2968/vshow}{GSS Data Explorer page for \code{adhdknow}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab never heard of it \tab not sure/dont know \tab no answer \tab other \tab don't know \tab Total \cr
   2002 \tab 1372 \tab 147 \tab 267 \tab 59 \tab 906 \tab 14 \tab 2765 \cr
   Total \tab 1372 \tab 147 \tab 267 \tab 59 \tab 906 \tab 14 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{mntlrel}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
