% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{adhdreal}
\alias{adhdreal}
\title{Is adhd a real disease}
\source{
General Social Survey https://gss.norc.org
}
\description{
adhdreal
}
\details{
Question Do you think ADHD is a real disease?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2970/vshow}{GSS Data Explorer page for \code{adhdreal}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{medsavtx}},
\code{\link{mntlrel}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
