% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_330.R
\name{adhosmed}
\alias{adhosmed}
\title{Would r have child be on med by hospital staff}
\source{
General Social Survey https://gss.norc.org
}
\description{
adhosmed
}
\details{
Question If (John/Mary) was your child, would you...
Check your child into a hospital?
A. If hospital staff said your child should be on medication, how likely would you be to do that?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2948/vshow}{GSS Data Explorer page for \code{adhosmed}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab Total \cr
   2002 \tab 2382 \tab 252 \tab 102 \tab 20 \tab 7 \tab 2 \tab 2765 \cr
   Total \tab 2382 \tab 252 \tab 102 \tab 20 \tab 7 \tab 2 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcoumed}},
\code{\link{adhos}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{grwingup}},
\code{\link{mntlill}}
}
\concept{Mental Health II}
\keyword{variable}
