% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{admirblk}
\alias{admirblk}
\title{R express admiration for blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
admirblk
}
\details{
Question Now, I would like to ask whether you have ever felt the following ways about blacks and their families. For each of the feelings that I ask you about, please tell me whether you have felt that way very often, fairly often, not too often, or never?
B. How often have you felt admiration for blacks?
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} fairly often
\item \verb{[3]} not too often
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1606/vshow}{GSS Data Explorer page for \code{admirblk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab fairly often \tab not too often \tab never \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 247 \tab 583 \tab 498 \tab 98 \tab 38 \tab 10 \tab 2992 \cr
   Total \tab 1518 \tab 247 \tab 583 \tab 498 \tab 98 \tab 38 \tab 10 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
