% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{adoption}
\alias{adoption}
\title{Best way to match birth mothers with adopters?}
\source{
General Social Survey https://gss.norc.org
}
\description{
adoption
}
\details{
Question As you know, when a parent is unable to care for a baby, the parent often puts the baby up for adoption. For certaIn kinds of babies, the number of couples who want to adopt is much greater than the number of babies available for adoption. What do you think is the best way to match birth mothers with couples who want to adopt?
}
\section{Values}{

\itemize{
\item \verb{[1]} a social service agency
\item \verb{[2]} go to the highest bidders
\item \verb{[3]} by lottery
\item \verb{[4]} birth mothers should decide
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1869/vshow}{GSS Data Explorer page for \code{adoption}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a social service agency \tab go to the highest bidders \tab by lottery \tab birth mothers should decide \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1460 \tab 758 \tab 32 \tab 104 \tab 477 \tab 67 \tab 6 \tab - \tab 2904 \cr
   Total \tab 1460 \tab 758 \tab 32 \tab 104 \tab 477 \tab 67 \tab 6 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{econsys}},
\code{\link{evbuyhme}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{fixer1}},
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{fixhome}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{haggle}},
\code{\link{homeyear}},
\code{\link{hrfambiz}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal3}},
\code{\link{inequal5}},
\code{\link{lawyer}},
\code{\link{lawyer1}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{layoffs}},
\code{\link{learnhme}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{natrecon}},
\code{\link{newowned}},
\code{\link{newused}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{realtor1}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{reldealr}},
\code{\link{relfixer}},
\code{\link{relhome}},
\code{\link{rellaw}},
\code{\link{relprivt}},
\code{\link{reqinfo}},
\code{\link{satcar}},
\code{\link{satfixes}},
\code{\link{sathome}},
\code{\link{satlawyr}},
\code{\link{sellauto}},
\code{\link{sellbaby}},
\code{\link{sellbed}},
\code{\link{sellhome}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{tagsales}},
\code{\link{transoth}},
\code{\link{transrel}},
\code{\link{typdealr}},
\code{\link{warrntyc}},
\code{\link{warrntyh}},
\code{\link{wholoand}},
\code{\link{whosold}},
\code{\link{wkfambiz}}

Other Children: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adoptkid}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{aidkids}},
\code{\link{allergic}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{cathsch}},
\code{\link{chembal}},
\code{\link{childs}},
\code{\link{chldben}},
\code{\link{chldcare}},
\code{\link{chldeduc}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldhrt}},
\code{\link{chldidel}},
\code{\link{chldidel_next}},
\code{\link{chldlove}},
\code{\link{chldmore}},
\code{\link{chldmorl}},
\code{\link{chldnum}},
\code{\link{chldpref_next}},
\code{\link{chldprob}},
\code{\link{chldsafe}},
\code{\link{chldsch}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldskls}},
\code{\link{chldsoon}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{chldvig}},
\code{\link{chrissch}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{comknows}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{discipln}},
\code{\link{eldersup}},
\code{\link{evkid}},
\code{\link{failure}},
\code{\link{famper4}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{fourkids}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{helpoth}},
\code{\link{hlth8}},
\code{\link{homesch}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{illphys}},
\code{\link{impkids}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpicked}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kid5up}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidhlth}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidlived}},
\code{\link{kidneedy}},
\code{\link{kidnofre}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidnum}},
\code{\link{kidout}},
\code{\link{kidpars}},
\code{\link{kidresp}},
\code{\link{kidskips}},
\code{\link{kidsocst}},
\code{\link{kidssol}},
\code{\link{kidtrble}},
\code{\link{kidxfilm}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{law7}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{mapaid}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{mntlill}},
\code{\link{morextrm}},
\code{\link{mykids}},
\code{\link{nextdoor}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obeythnk}},
\code{\link{obrespct}},
\code{\link{onekid}},
\code{\link{oth3}},
\code{\link{othkids}},
\code{\link{othsch}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{popular}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{raised}},
\code{\link{relsch}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{sellbaby}},
\code{\link{spanking}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spendeve}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}},
\code{\link{spwrkpar}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{stressfl}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{sufadult}},
\code{\link{thnkself}},
\code{\link{threkids}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{trbllaw}},
\code{\link{twokids}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{workhard}},
\code{\link{zombies}}

Other Purchases/Transactions: 
\code{\link{borrowed}},
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{evbuyhme}},
\code{\link{famfinan}},
\code{\link{fixer1}},
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{fixhome}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{haggle}},
\code{\link{homeless}},
\code{\link{homeyear}},
\code{\link{lawyer}},
\code{\link{lawyer1}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{learnhme}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{newowned}},
\code{\link{newused}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{realtor1}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{reldealr}},
\code{\link{relfixer}},
\code{\link{relhome}},
\code{\link{rellaw}},
\code{\link{relprivt}},
\code{\link{satcar}},
\code{\link{satfixes}},
\code{\link{sathome}},
\code{\link{satlawyr}},
\code{\link{sellauto}},
\code{\link{sellbaby}},
\code{\link{sellbed}},
\code{\link{sellhome}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{tagsales}},
\code{\link{transoth}},
\code{\link{transrel}},
\code{\link{typdealr}},
\code{\link{warrntyc}},
\code{\link{warrntyh}},
\code{\link{wholoand}},
\code{\link{whosold}}
}
\concept{Children}
\concept{Markets}
\concept{Purchases/Transactions}
\keyword{variable}
