% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{adoptkid}
\alias{adoptkid}
\title{Have ever legally adopted a child}
\source{
General Social Survey https://gss.norc.org
}
\description{
adoptkid
}
\details{
Question Have you ever legally adopted a child?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3914/vshow}{GSS Data Explorer page for \code{adoptkid}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{toldsxor}}

Other Children: 
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}}

Other Family: 
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{marunion}}
}
\concept{Children}
\concept{Family}
\concept{Sexual Orientation}
\keyword{variable}
