% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{adoptus}
\alias{adoptus}
\title{Immigrants living in us should adopt us culture}
\source{
General Social Survey https://gss.norc.org
}
\description{
adoptus
}
\details{
Question Do you think immigrants living in the U.S. should adopt U.S. culture, keep their own culture, or do both?
}
\section{Values}{

\itemize{
\item \verb{[1]} adopt u.s. culture
\item \verb{[2]} keep their own culture
\item \verb{[3]} do both
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7755/vshow}{GSS Data Explorer page for \code{adoptus}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab adopt u.s. culture \tab keep their own culture \tab do both \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 258 \tab 195 \tab 1860 \tab 18 \tab 1203 \tab 6 \tab 4 \tab 3544 \cr
   2024 \tab - \tab 273 \tab 188 \tab 1738 \tab 30 \tab 1067 \tab 5 \tab 8 \tab 3309 \cr
   Total \tab 0 \tab 531 \tab 383 \tab 3598 \tab 48 \tab 2270 \tab 11 \tab 12 \tab 6853 \cr
}
}

\concept{Core}
\keyword{variable}
