% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_330.R
\name{adpsy}
\alias{adpsy}
\title{Would r go to a psychiatrist if child was r's}
\source{
General Social Survey https://gss.norc.org
}
\description{
adpsy
}
\details{
Question If (John/Mary) was your child, would you...
Go to a psychiatrist for help?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2945/vshow}{GSS Data Explorer page for \code{adpsy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2002 \tab 1372 \tab 741 \tab 605 \tab 36 \tab 11 \tab - \tab - \tab 2765 \cr
   2024 \tab 1661 \tab 1189 \tab 406 \tab 27 \tab 5 \tab - \tab 21 \tab 3309 \cr
   Total \tab 3033 \tab 1930 \tab 1011 \tab 63 \tab 16 \tab 0 \tab 21 \tab 6074 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcoumed}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsymed}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{grwingup}},
\code{\link{mntlill}}

Other Children: 
\code{\link{adcoumed}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsymed}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{grwingup}},
\code{\link{mntlill}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
