% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{adults_exp}
\alias{adults_exp}
\title{Adults 18+ in hh(experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
adults_exp
}
\details{
Question Including yourself, how many individuals 18 or older live in this household?
If you have roommates, please include them. Please also include those who usually live here but are away temporarily, such as those on business trips, at school, or temporarily in a hospital.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 adult in household
\item \verb{[8]} 8 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7983/vshow}{GSS Data Explorer page for \code{adults_exp}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab 1 adult in household \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 or more \tab iap \tab no answer \tab skipped on web \tab don't know \tab Total \cr
   2022 \tab - \tab 699 \tab 764 \tab 171 \tab 63 \tab 12 \tab 6 \tab 1 \tab 1 \tab 1788 \tab 4 \tab 35 \tab - \tab 3544 \cr
   2024 \tab - \tab 622 \tab 681 \tab 178 \tab 65 \tab 24 \tab 2 \tab - \tab 4 \tab 1684 \tab 3 \tab 44 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 1321 \tab 1445 \tab 349 \tab 128 \tab 36 \tab 8 \tab 1 \tab 5 \tab 3472 \tab 7 \tab 79 \tab 2 \tab 6853 \cr
}
}

\keyword{variable}
