% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{adultsinhh}
\alias{adultsinhh}
\title{Total adults in hh (household composition version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
adultsinhh
}
\details{
Question Calculated variable: Total adults in household according to roster
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 adult in household
\item \verb{[8]} 8 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7982/vshow}{GSS Data Explorer page for \code{adultsinhh}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab 1 adult in household \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 or more \tab iap \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab - \tab 1188 \tab 2004 \tab 441 \tab 133 \tab 47 \tab 4 \tab 2 \tab 2 \tab 203 \tab 8 \tab - \tab 4032 \cr
   2022 \tab - \tab 673 \tab 818 \tab 167 \tab 78 \tab 15 \tab 4 \tab - \tab - \tab 1764 \tab 4 \tab 21 \tab 3544 \cr
   2024 \tab - \tab 670 \tab 772 \tab 164 \tab 59 \tab 10 \tab - \tab 1 \tab - \tab 1630 \tab 2 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 2531 \tab 3594 \tab 772 \tab 270 \tab 72 \tab 8 \tab 3 \tab 2 \tab 3597 \tab 14 \tab 22 \tab 10885 \cr
}
}

\keyword{variable}
