% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{advances}
\alias{advances}
\title{Have you advanced, same level or lost}
\source{
General Social Survey https://gss.norc.org
}
\description{
advances
}
\details{
Question Since your first full-time job with this organization, have you advanced rapidly, made steady advances, stayed at about the same level or lost some ground?
}
\section{Values}{

\itemize{
\item \verb{[1]} advanced rapidly
\item \verb{[2]} steady advance
\item \verb{[3]} samelevel
\item \verb{[4]} lost some ground
\item \verb{[7]} 1st job with org
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1306/vshow}{GSS Data Explorer page for \code{advances}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab advanced rapidly \tab steady advance \tab samelevel \tab lost some ground \tab 1st job with org \tab no answer \tab Total \cr
   1991 \tab 605 \tab 80 \tab 314 \tab 397 \tab 23 \tab 64 \tab 34 \tab 1517 \cr
   Total \tab 605 \tab 80 \tab 314 \tab 397 \tab 23 \tab 64 \tab 34 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}
}
\concept{Work Organizations}
\keyword{variable}
