% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{advantge}
\alias{advantge}
\title{Other ppl will take advantage of r if not careful}
\source{
General Social Survey https://gss.norc.org
}
\description{
advantge
}
\details{
Question To what extent do you agree or disagree with the following statements?
C. If you are not careful, other people will take advantage of you
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4238/vshow}{GSS Data Explorer page for \code{advantge}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   2002 \tab 1616 \tab 280 \tab 514 \tab 180 \tab 143 \tab 12 \tab 9 \tab 11 \tab - \tab 2765 \cr
   Total \tab 1616 \tab 280 \tab 514 \tab 180 \tab 143 \tab 12 \tab 9 \tab 11 \tab 0 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{aidkids}},
\code{\link{bestfrd}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{comyear}},
\code{\link{cousins}},
\code{\link{cowrkfrd}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{hapunhap}},
\code{\link{helpfrds}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{inlawnum}},
\code{\link{kidcall}},
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidpars}},
\code{\link{kidvisit}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{localgvt}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{malive}},
\code{\link{marcohab}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{neifrd}},
\code{\link{niecenep}},
\code{\link{othfrd}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{palive}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{peocntct}},
\code{\link{posslq}},
\code{\link{posslqy}},
\code{\link{relcall}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{talkedto}},
\code{\link{trustpeo}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{usefrds}},
\code{\link{wantbest}}

Other People: 
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{fair}},
\code{\link{fair5}},
\code{\link{helpful}},
\code{\link{trust}},
\code{\link{trust5}},
\code{\link{trusting}},
\code{\link{trustpeo}},
\code{\link{trusty}},
\code{\link{twoclass}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{People}
\keyword{variable}
