% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_275.R
\name{advfam}
\alias{advfam}
\title{R can ask other family members software help}
\source{
General Social Survey https://gss.norc.org
}
\description{
advfam
}
\details{
Question Do the people you can ask for advice include? (Name as many as are correct):
E. Other family?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2503/vshow}{GSS Data Explorer page for \code{advfam}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2145 \tab 185 \tab 392 \tab 95 \tab - \tab 2817 \cr
   2002 \tab 2142 \tab 200 \tab 375 \tab 47 \tab 1 \tab 2765 \cr
   Total \tab 4287 \tab 385 \tab 767 \tab 142 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}}
}
\concept{Information Society}
\keyword{variable}
