% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_634.R
\name{advfmsci}
\alias{advfmsci}
\title{All of the greatest advances for humanity have come from science and technology}
\source{
General Social Survey https://gss.norc.org
}
\description{
advfmsci
}
\details{
Question All of the greatest advances for humanity have come from science and technology.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7541/vshow}{GSS Data Explorer page for \code{advfmsci}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 422 \tab 366 \tab 1135 \tab 1271 \tab 652 \tab 134 \tab 2 \tab 3 \tab 47 \tab 4032 \cr
   Total \tab 422 \tab 366 \tab 1135 \tab 1271 \tab 652 \tab 134 \tab 2 \tab 3 \tab 47 \tab 4032 \cr
}
}

\keyword{variable}
