% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{advfront}
\alias{advfront}
\title{Sci rsch is necessary and should be supported by federal govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
advfront
}
\details{
Question I’m going to read to you some statements like those you might find in a newspaper or magazine article. For each statement, please tell me if you strongly agree, agree, disagree, or strongly disagree.
D. Even if it brings no immediate benefits, scientific research that advances the frontiers of knowledge is necessary and should be supported by the federal government.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3319/vshow}{GSS Data Explorer page for \code{advfront}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}

Other GSS Next: 
\code{\link{nextgen}},
\code{\link{toofast}}

Other Federal Government: 
\code{\link{leadsci}}

Other Science: 
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}
}
\concept{Federal Government}
\concept{GSS Next}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
