% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_101.R
\name{advisor4}
\alias{advisor4}
\title{Person #4 is r's advisor}
\source{
General Social Survey https://gss.norc.org
}
\description{
advisor4
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
J4. Fourth Person is Advisor
}
\section{Values}{

\itemize{
\item \code{1} mentioned
\item \code{2} not mentioned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/923/vshow}{GSS Data Explorer page for \code{advisor4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab Total \cr
   1985 \tab 920 \tab 86 \tab 525 \tab 3 \tab 1534 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab 1819 \cr
   2004 \tab 2472 \tab 59 \tab 235 \tab 46 \tab 2812 \cr
   Total \tab 5211 \tab 145 \tab 760 \tab 49 \tab 6165 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor5}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}}
}
\concept{Social Networks}
\keyword{variable}
