% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{advofrnd}
\alias{advofrnd}
\title{R can ask acquaintance software help}
\source{
General Social Survey https://gss.norc.org
}
\description{
advofrnd
}
\details{
Question Do the people you can ask for advice include? (Name as many as are correct):
K. Other friend or acquaintance?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2509/vshow}{GSS Data Explorer page for \code{advofrnd}}.
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
