% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_275.R
\name{advpar}
\alias{advpar}
\title{R can ask parent for software help}
\source{
General Social Survey https://gss.norc.org
}
\description{
advpar
}
\details{
Question Do the people you can ask for advice include? (Name as many as are correct):
C. Mother or father?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2501/vshow}{GSS Data Explorer page for \code{advpar}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2145 \tab 100 \tab 475 \tab 97 \tab 2817 \cr
   2002 \tab 2142 \tab 153 \tab 423 \tab 47 \tab 2765 \cr
   Total \tab 4287 \tab 253 \tab 898 \tab 144 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}}
}
\concept{Information Society}
\keyword{variable}
