% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_304.R
\name{advsched}
\alias{advsched}
\title{How far in advance do you schedule work}
\source{
General Social Survey https://gss.norc.org
}
\description{
advsched
}
\details{
Question How far in advance do you usually know what days and hours you will need to work?
(HAND CARD L2)
}
\section{Values}{

\itemize{
\item \code{1} one day or less in advance
\item \code{2} 2 to 3 days in advance
\item \code{3} 4 to 7 days in advance
\item \code{4} between 1 and 2 weeks in advance
\item \code{5} between 3 and 4 weeks in advance
\item \code{6} more than 4 weeks in advance
\item \code{7} my schedule never changes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6296/vshow}{GSS Data Explorer page for \code{advsched}}.
}

\seealso{
Other Flexible Work: 
\code{\link{leasthrs}},
\code{\link{sethrs}}
}
\concept{Flexible Work}
\keyword{variable}
