% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_275.R
\name{advstu}
\alias{advstu}
\title{R can ask other student software help}
\source{
General Social Survey https://gss.norc.org
}
\description{
advstu
}
\details{
Question Do the people you can ask for advice include? (Name as many as are correct):
G. A fellow student?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2505/vshow}{GSS Data Explorer page for \code{advstu}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2145 \tab 117 \tab 457 \tab 98 \tab 2817 \cr
   2002 \tab 2142 \tab 109 \tab 467 \tab 47 \tab 2765 \cr
   Total \tab 4287 \tab 226 \tab 924 \tab 145 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsup}},
\code{\link{advtchr}}
}
\concept{Information Society}
\keyword{variable}
