% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_275.R
\name{advsup}
\alias{advsup}
\title{R can ask work supervisor software help}
\source{
General Social Survey https://gss.norc.org
}
\description{
advsup
}
\details{
Question Do the people you can ask for advice include? (Name as many as are correct):
H. A supervisor or trainer at work?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2506/vshow}{GSS Data Explorer page for \code{advsup}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2145 \tab 254 \tab 323 \tab 95 \tab 2817 \cr
   2002 \tab 2142 \tab 240 \tab 336 \tab 47 \tab 2765 \cr
   Total \tab 4287 \tab 494 \tab 659 \tab 142 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advstu}},
\code{\link{advtchr}}
}
\concept{Information Society}
\keyword{variable}
