% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{affctlaw}
\alias{affctlaw}
\title{How likely congress give serious attention to r's demands}
\source{
General Social Survey https://gss.norc.org
}
\description{
affctlaw
}
\details{
Question If you made such an effort, how likely is it that the Congress would give serious attention to your demands?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} fairly likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4943/vshow}{GSS Data Explorer page for \code{affctlaw}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polgreed}},
\code{\link{polint1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
