% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{affrmact}
\alias{affrmact}
\title{Favor preference in hiring black people}
\source{
General Social Survey https://gss.norc.org
}
\description{
affrmact
}
\details{
Question A. Some people say that because of past discrimination, Black people should be given preference in hiring and promotion. Others say that such preference in hiring and promotion of Black people is wrong because it discriminates against whites. What about your opinion -- are you for or against preferential hiring and promotion of Black people? IF FAVORS: A. Do you favor preference in hiring and promotion strongly or not strongly? IF OPPOSES: B. Do you oppose preference in hiring and promotion strongly or not strongly?
}
\section{Values}{

\itemize{
\item \code{1} strongly favors
\item \code{2} not strongly favors
\item \code{3} not strongly opposes
\item \code{4} strongly opposes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
   2004 \tab A/B \tab partial \cr
   2006 \tab A/B \tab partial \cr
   2008 \tab A/B \tab partial \cr
   2010 \tab A/B \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/423/vshow}{GSS Data Explorer page for \code{affrmact}}.
}

\seealso{
Other Core: 
\code{\link{blksimp}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}
}
\concept{Affirmative Action}
\concept{Core}
\concept{Federal Government}
\keyword{variable}
