% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{afterlif}
\alias{afterlif}
\title{Belief in life after death}
\source{
General Social Survey https://gss.norc.org
}
\description{
afterlif
}
\details{
Question Do you believe in. . .
A. Life after death
}
\section{Values}{

\itemize{
\item \code{1} yes, definitely
\item \code{2} yes, probably
\item \code{3} no, probably not
\item \code{4} no, definitely not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4651/vshow}{GSS Data Explorer page for \code{afterlif}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably not \tab no, definitely not \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 674 \tab 278 \tab 150 \tab 123 \tab 94 \tab 40 \tab 1517 \cr
   1998 \tab 1548 \tab 660 \tab 247 \tab 110 \tab 110 \tab 110 \tab 47 \tab 2832 \cr
   2008 \tab 658 \tab 745 \tab 279 \tab 152 \tab 127 \tab 54 \tab 8 \tab 2023 \cr
   2018 \tab 1173 \tab 662 \tab 242 \tab 126 \tab 99 \tab 36 \tab 10 \tab 2348 \cr
   Total \tab 3537 \tab 2741 \tab 1046 \tab 538 \tab 459 \tab 294 \tab 105 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
