% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{agape1}
\alias{agape1}
\title{Would suffer than let the one r love suffer}
\source{
General Social Survey https://gss.norc.org
}
\description{
agape1
}
\details{
Question Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or strongly disagree with the following statements?
A. I would rather suffer myself than let the one I love suffer.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree somewhat
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3053/vshow}{GSS Data Explorer page for \code{agape1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab neither agree nor disagree \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 812 \tab 364 \tab 100 \tab 41 \tab 13 \tab 5 \tab 5 \tab 2812 \cr
   Total \tab 1472 \tab 812 \tab 364 \tab 100 \tab 41 \tab 13 \tab 5 \tab 5 \tab 2812 \cr
}
}

\concept{Altruistic Love}
\keyword{variable}
