% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{agape4}
\alias{agape4}
\title{Would endure all things for the one r love}
\source{
General Social Survey https://gss.norc.org
}
\description{
agape4
}
\details{
Question Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or strongly disagree with the following statements?
D. I would endure all things for the sake of the one I love.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree somewhat
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3056/vshow}{GSS Data Explorer page for \code{agape4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab neither agree nor disagree \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 557 \tab 475 \tab 146 \tab 106 \tab 45 \tab 5 \tab 6 \tab 2812 \cr
   Total \tab 1472 \tab 557 \tab 475 \tab 146 \tab 106 \tab 45 \tab 5 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Altruistic Love: 
\code{\link{agape2}},
\code{\link{agape3}}
}
\concept{Altruistic Love}
\keyword{variable}
