% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{age2}
\alias{age2}
\title{Age of person #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
age2
}
\details{
Question How old is (NAME)?  PROBE:  Your best guess. ASK FOR EACH NAME.
2. Age of Second Person
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/946/vshow}{GSS Data Explorer page for \code{age2}}.
}

\seealso{
Other Social Networks: 
\code{\link{age1}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}}

Other Social Networks: 
\code{\link{age1}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}}
}
\concept{Social Networks}
\keyword{variable}
