% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{age3}
\alias{age3}
\title{Age of person #3}
\source{
General Social Survey https://gss.norc.org
}
\description{
age3
}
\details{
Question How old is (NAME)?  PROBE:  Your best guess. ASK FOR EACH NAME.
3. Age of Third Person
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/947/vshow}{GSS Data Explorer page for \code{age3}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 11 \tab 12 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 77 \tab 79 \tab 80 \tab 81 \tab 82 \tab 85 \tab 86 \tab 87 \tab 89 \tab don't know \tab no answer \tab 14 \tab 76 \tab 78 \tab 8 \tab 83 \tab 84 \tab 88 \tab 90 \tab 92 \tab Total \cr
   1985 \tab 599 \tab 1 \tab 1 \tab 4 \tab 6 \tab 7 \tab 6 \tab 7 \tab 15 \tab 15 \tab 10 \tab 17 \tab 14 \tab 14 \tab 24 \tab 29 \tab 19 \tab 18 \tab 34 \tab 24 \tab 28 \tab 18 \tab 19 \tab 30 \tab 21 \tab 20 \tab 27 \tab 15 \tab 21 \tab 21 \tab 33 \tab 20 \tab 9 \tab 37 \tab 14 \tab 12 \tab 12 \tab 17 \tab 26 \tab 12 \tab 9 \tab 13 \tab 7 \tab 18 \tab 15 \tab 11 \tab 16 \tab 15 \tab 24 \tab 6 \tab 7 \tab 6 \tab 11 \tab 18 \tab 6 \tab 6 \tab 6 \tab 2 \tab 10 \tab 4 \tab 3 \tab 4 \tab 4 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 12 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   2004 \tab 2240 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 5 \tab 3 \tab 11 \tab 8 \tab 10 \tab 6 \tab 11 \tab 8 \tab 11 \tab 7 \tab 9 \tab 8 \tab 12 \tab 9 \tab 9 \tab 11 \tab 14 \tab 9 \tab 12 \tab 10 \tab 20 \tab 8 \tab 14 \tab 9 \tab 7 \tab 21 \tab 7 \tab 6 \tab 17 \tab 10 \tab 17 \tab 8 \tab 9 \tab 12 \tab 14 \tab 16 \tab 17 \tab 4 \tab 9 \tab 6 \tab 13 \tab 5 \tab 9 \tab 5 \tab 4 \tab 6 \tab - \tab 3 \tab 5 \tab 1 \tab 6 \tab 5 \tab 6 \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 50 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2812 \cr
   Total \tab 2839 \tab 2 \tab 1 \tab 5 \tab 9 \tab 8 \tab 8 \tab 9 \tab 20 \tab 18 \tab 21 \tab 25 \tab 24 \tab 20 \tab 35 \tab 37 \tab 30 \tab 25 \tab 43 \tab 32 \tab 40 \tab 27 \tab 28 \tab 41 \tab 35 \tab 29 \tab 39 \tab 25 \tab 41 \tab 29 \tab 47 \tab 29 \tab 16 \tab 58 \tab 21 \tab 18 \tab 29 \tab 27 \tab 43 \tab 20 \tab 18 \tab 25 \tab 21 \tab 34 \tab 32 \tab 15 \tab 25 \tab 21 \tab 37 \tab 11 \tab 16 \tab 11 \tab 15 \tab 24 \tab 6 \tab 9 \tab 11 \tab 3 \tab 16 \tab 9 \tab 9 \tab 5 \tab 7 \tab 7 \tab 2 \tab 5 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 13 \tab 58 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}}
}
\concept{Social Networks}
\keyword{variable}
