% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{agecmeus}
\alias{agecmeus}
\title{If born outside usa, what age were you when you permanently moved to usa}
\source{
General Social Survey https://gss.norc.org
}
\description{
agecmeus
}
\details{
Question If you were born outside US, what age were you when you permanently moved to the US?
}
\section{Values}{

\itemize{
\item \verb{[96]} have not permanently moved to us
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3899/vshow}{GSS Data Explorer page for \code{agecmeus}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 46 \tab 5 \tab 52 \tab 56 \tab 6 \tab 7 \tab 9 \tab have not permanently moved to us \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 1838 \tab 9 \tab 5 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 4 \tab 7 \tab 10 \tab 8 \tab 9 \tab 9 \tab 1 \tab 7 \tab 5 \tab 7 \tab 7 \tab 7 \tab 3 \tab 3 \tab 2 \tab 4 \tab 9 \tab 2 \tab 6 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 8 \tab 6 \tab - \tab 2023 \cr
   Total \tab 1838 \tab 9 \tab 5 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 4 \tab 7 \tab 10 \tab 8 \tab 9 \tab 9 \tab 1 \tab 7 \tab 5 \tab 7 \tab 7 \tab 7 \tab 3 \tab 3 \tab 2 \tab 4 \tab 9 \tab 2 \tab 6 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 8 \tab 6 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{bornsp}},
\code{\link{curempyr}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{natdeg}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{typevisa}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work3yrs}}

Other Age: 
\code{\link{age}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{cohort}}

Other Citizenship: 
\code{\link{actassoc}},
\code{\link{bornsp}},
\code{\link{buypol}},
\code{\link{fucitzn}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obeylaws}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{typevisa}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{visa}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Immigration: 
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Residence: 
\code{\link{abroad}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{comamind}},
\code{\link{comtype}},
\code{\link{comyear}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwell5}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{geomobil}},
\code{\link{hspasn10}},
\code{\link{liveasns}},
\code{\link{livecom}},
\code{\link{livecom1}},
\code{\link{livehome}},
\code{\link{livehsps}},
\code{\link{loclived}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{res16}},
\code{\link{res161}},
\code{\link{res2006}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Age}
\concept{Citizenship}
\concept{Global Economics}
\concept{Immigration}
\concept{Residence}
\keyword{variable}
