% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{agedchld}
\alias{agedchld}
\title{Should grown children live with parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
agedchld
}
\details{
Question As you know, many grown children move back home to live with their parents. Do you think this is generally a good idea or a bad idea?
}
\section{Values}{

\itemize{
\item \verb{[1]} a good idea
\item \verb{[2]} a bad idea
\item \verb{[3]} depends
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6175/vshow}{GSS Data Explorer page for \code{agedchld}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a good idea \tab a bad idea \tab depends \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2414 \tab 192 \tab 153 \tab 102 \tab 4 \tab 2 \tab 2867 \cr
   Total \tab 2414 \tab 192 \tab 153 \tab 102 \tab 4 \tab 2 \tab 2867 \cr
}
}

\seealso{
Other Core: 
\code{\link{aged}},
\code{\link{agedpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{partfull}},
\code{\link{smoke}},
\code{\link{socpars}},
\code{\link{socsibs}},
\code{\link{weekswrk}}
}
\concept{Core}
\keyword{variable}
