% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{agedv}
\alias{agedv}
\title{Should aged live with their children (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
agedv
}
\details{
Question As you know, many older people share a home with their grown children. Do you think this is generally a good idea or a bad idea?
}
\section{Values}{

\itemize{
\item \verb{[1]} a good idea
\item \verb{[2]} a bad idea
\item \verb{[3]} depends (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7594/vshow}{GSS Data Explorer page for \code{agedv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a good idea \tab a bad idea \tab depends (phone mode only: volunteered) \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2673 \tab 359 \tab 167 \tab 831 \tab 2 \tab - \tab 4032 \cr
   2022 \tab 2961 \tab 144 \tab 58 \tab 379 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 5634 \tab 503 \tab 225 \tab 1210 \tab 3 \tab 1 \tab 7576 \cr
}
}

\keyword{variable}
