% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_668.R
\name{agehef7}
\alias{agehef7}
\title{Age of seventh person in household screening (if applicable)}
\source{
General Social Survey https://gss.norc.org
}
\description{
agehef7
}
\details{
Question How old was (NAME) on their last birthday?
Was (NAME) over or under 18 on their last birthday?
}
\section{Values}{

\itemize{
\item \code{1} 18-29
\item \code{2} 30-39
\item \code{3} 40-49
\item \code{4} 50-64
\item \code{5} 65 or older
\item \code{6} 18 or older, age unknown
\item \code{7} under 18
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7995/vshow}{GSS Data Explorer page for \code{agehef7}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab 18-29 \tab 99 \tab Total \cr
   2022 \tab 3540 \tab 3 \tab 1 \tab 3544 \cr
   Total \tab 3540 \tab 3 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
