% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{agehef8}
\alias{agehef8}
\title{Age of eighth person in household screening (if applicable)}
\source{
General Social Survey https://gss.norc.org
}
\description{
agehef8
}
\details{
Question How old was (NAME) on their last birthday?
Was (NAME) over or under 18 on their last birthday?
}
\section{Values}{

\itemize{
\item \verb{[1]} 18-29
\item \verb{[2]} 30-39
\item \verb{[3]} 40-49
\item \verb{[4]} 50-64
\item \verb{[5]} 65 or older
\item \verb{[6]} 18 or older, age unknown
\item \verb{[7]} under 18
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7996/vshow}{GSS Data Explorer page for \code{agehef8}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab 18-29 \tab 99 \tab Total \cr
   2022 \tab 3541 \tab 2 \tab 1 \tab 3544 \cr
   Total \tab 3541 \tab 2 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
