% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{agekdbrn}
\alias{agekdbrn}
\title{R's age when 1st child born}
\source{
General Social Survey https://gss.norc.org
}
\description{
agekdbrn
}
\details{
Question How old were you when your first child was born?
As of 2024, AGEKDBRN has been recoded to only account for first child born during ages 18-45. This has been retroactively applied to all previous years of AGEKDBRN. The original AGEKDBRN containing data from 1994 to 2022 has been renamed AGEKDBRN_9422.
}
\section{Values}{

\itemize{
\item \verb{[17]} 17 and younger
\item \verb{[45]} 45 and older
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/54/vshow}{GSS Data Explorer page for \code{agekdbrn}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 17 and younger \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 45 and older \tab don't know \tab iap \tab no answer \tab 44 \tab 43 \tab skipped on web \tab Total \cr
   1994 \tab - \tab 119 \tab 109 \tab 125 \tab 142 \tab 128 \tab 101 \tab 106 \tab 114 \tab 117 \tab 92 \tab 91 \tab 82 \tab 61 \tab 57 \tab 39 \tab 28 \tab 18 \tab 17 \tab 14 \tab 10 \tab 3 \tab 6 \tab 6 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1377 \tab 19 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 175 \tab 133 \tab 151 \tab 182 \tab 203 \tab 170 \tab 135 \tab 130 \tab 139 \tab 121 \tab 103 \tab 82 \tab 53 \tab 64 \tab 39 \tab 32 \tab 28 \tab 19 \tab 26 \tab 12 \tab 13 \tab 11 \tab 2 \tab 6 \tab 4 \tab - \tab 8 \tab 10 \tab 824 \tab 28 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 179 \tab 124 \tab 176 \tab 166 \tab 177 \tab 144 \tab 156 \tab 111 \tab 151 \tab 92 \tab 89 \tab 93 \tab 66 \tab 70 \tab 36 \tab 52 \tab 22 \tab 19 \tab 18 \tab 15 \tab 9 \tab 5 \tab 9 \tab 3 \tab 7 \tab 6 \tab 3 \tab 3 \tab 802 \tab 28 \tab - \tab 1 \tab - \tab 2832 \cr
   2000 \tab - \tab 177 \tab 142 \tab 154 \tab 164 \tab 164 \tab 162 \tab 118 \tab 130 \tab 140 \tab 97 \tab 96 \tab 78 \tab 74 \tab 64 \tab 37 \tab 38 \tab 31 \tab 21 \tab 24 \tab 12 \tab 8 \tab 10 \tab 8 \tab 4 \tab 4 \tab 4 \tab 2 \tab 7 \tab 813 \tab 32 \tab 1 \tab 1 \tab - \tab 2817 \cr
   2002 \tab - \tab 148 \tab 122 \tab 158 \tab 180 \tab 187 \tab 156 \tab 134 \tab 127 \tab 108 \tab 94 \tab 96 \tab 78 \tab 62 \tab 69 \tab 45 \tab 47 \tab 39 \tab 18 \tab 25 \tab 21 \tab 15 \tab 5 \tab 5 \tab 4 \tab 3 \tab 3 \tab 3 \tab 2 \tab 799 \tab 9 \tab 1 \tab 2 \tab - \tab 2765 \cr
   2004 \tab - \tab 141 \tab 132 \tab 168 \tab 177 \tab 181 \tab 157 \tab 115 \tab 131 \tab 114 \tab 104 \tab 119 \tab 86 \tab 56 \tab 87 \tab 52 \tab 55 \tab 33 \tab 27 \tab 26 \tab 12 \tab 12 \tab 10 \tab 11 \tab 6 \tab 3 \tab 2 \tab 6 \tab 10 \tab 767 \tab 9 \tab 1 \tab 2 \tab - \tab 2812 \cr
   2006 \tab - \tab 183 \tab 134 \tab 156 \tab 201 \tab 216 \tab 148 \tab 149 \tab 120 \tab 132 \tab 122 \tab 91 \tab 80 \tab 84 \tab 79 \tab 42 \tab 49 \tab 31 \tab 21 \tab 21 \tab 23 \tab 15 \tab 13 \tab 6 \tab 10 \tab 4 \tab 5 \tab 8 \tab 1 \tab 2348 \tab 13 \tab 2 \tab 3 \tab - \tab 4510 \cr
   2008 \tab - \tab 125 \tab 102 \tab 104 \tab 135 \tab 149 \tab 115 \tab 92 \tab 79 \tab 93 \tab 88 \tab 71 \tab 67 \tab 46 \tab 57 \tab 26 \tab 30 \tab 19 \tab 19 \tab 26 \tab 13 \tab 7 \tab 11 \tab 5 \tab 2 \tab - \tab 4 \tab 2 \tab 6 \tab 521 \tab 7 \tab 2 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 145 \tab 117 \tab 110 \tab 125 \tab 133 \tab 95 \tab 87 \tab 88 \tab 69 \tab 64 \tab 70 \tab 65 \tab 48 \tab 52 \tab 42 \tab 33 \tab 23 \tab 15 \tab 16 \tab 19 \tab 13 \tab 6 \tab 6 \tab 6 \tab 6 \tab 6 \tab 8 \tab 2 \tab 562 \tab 10 \tab 2 \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab 126 \tab 101 \tab 102 \tab 110 \tab 134 \tab 78 \tab 95 \tab 82 \tab 94 \tab 70 \tab 78 \tab 58 \tab 52 \tab 55 \tab 32 \tab 32 \tab 28 \tab 29 \tab 21 \tab 13 \tab 11 \tab 10 \tab 7 \tab 6 \tab 3 \tab 1 \tab 2 \tab 3 \tab 536 \tab 5 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 167 \tab 117 \tab 135 \tab 134 \tab 168 \tab 120 \tab 114 \tab 107 \tab 123 \tab 70 \tab 62 \tab 63 \tab 60 \tab 86 \tab 37 \tab 51 \tab 38 \tab 27 \tab 40 \tab 21 \tab 16 \tab 14 \tab 10 \tab 11 \tab 6 \tab 5 \tab 10 \tab 7 \tab 704 \tab 10 \tab 1 \tab 4 \tab - \tab 2538 \cr
   2016 \tab - \tab 185 \tab 135 \tab 165 \tab 164 \tab 170 \tab 115 \tab 124 \tab 101 \tab 144 \tab 109 \tab 89 \tab 76 \tab 71 \tab 97 \tab 50 \tab 55 \tab 41 \tab 31 \tab 38 \tab 19 \tab 18 \tab 13 \tab 13 \tab 15 \tab 4 \tab 5 \tab 6 \tab 3 \tab 797 \tab 11 \tab 1 \tab 2 \tab - \tab 2867 \cr
   2018 \tab - \tab 134 \tab 96 \tab 131 \tab 126 \tab 142 \tab 113 \tab 101 \tab 115 \tab 120 \tab 63 \tab 67 \tab 87 \tab 59 \tab 83 \tab 37 \tab 43 \tab 38 \tab 20 \tab 28 \tab 9 \tab 10 \tab 10 \tab 7 \tab 10 \tab 4 \tab 3 \tab 7 \tab 10 \tab 663 \tab 9 \tab 1 \tab 2 \tab - \tab 2348 \cr
   2021 \tab - \tab 177 \tab 132 \tab 157 \tab 213 \tab 207 \tab 158 \tab 170 \tab 171 \tab 177 \tab 134 \tab 143 \tab 144 \tab 109 \tab 147 \tab 72 \tab 94 \tab 80 \tab 70 \tab 58 \tab 41 \tab 29 \tab 37 \tab 29 \tab 20 \tab 9 \tab 5 \tab 16 \tab - \tab 1212 \tab - \tab 2 \tab 2 \tab 17 \tab 4032 \cr
   2022 \tab - \tab 186 \tab 133 \tab 168 \tab 176 \tab 207 \tab 147 \tab 131 \tab 129 \tab 145 \tab 138 \tab 122 \tab 117 \tab 91 \tab 121 \tab 46 \tab 59 \tab 52 \tab 42 \tab 45 \tab 28 \tab 24 \tab 24 \tab 22 \tab 14 \tab 9 \tab 11 \tab 11 \tab 1 \tab 1107 \tab 8 \tab 2 \tab 6 \tab 22 \tab 3544 \cr
   2024 \tab - \tab 198 \tab 121 \tab 144 \tab 144 \tab 176 \tab 134 \tab 147 \tab 119 \tab 128 \tab 108 \tab 101 \tab 115 \tab 86 \tab 103 \tab 78 \tab 72 \tab 43 \tab 44 \tab 43 \tab 27 \tab 16 \tab 23 \tab 13 \tab 22 \tab 7 \tab 7 \tab 11 \tab 6 \tab 1029 \tab 5 \tab 6 \tab 4 \tab 29 \tab 3309 \cr
   Total \tab 0 \tab 2565 \tab 1950 \tab 2304 \tab 2539 \tab 2742 \tab 2113 \tab 1974 \tab 1854 \tab 1994 \tab 1566 \tab 1488 \tab 1371 \tab 1078 \tab 1291 \tab 710 \tab 770 \tab 564 \tab 439 \tab 469 \tab 295 \tab 219 \tab 208 \tab 159 \tab 145 \tab 74 \tab 69 \tab 104 \tab 72 \tab 14861 \tab 203 \tab 23 \tab 30 \tab 68 \tab 46311 \cr
}
}

\seealso{
Other Core: 
\code{\link{age}},
\code{\link{codeg}},
\code{\link{coeduc}},
\code{\link{degree}},
\code{\link{educ}},
\code{\link{maeduc}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{speduc}}

Other Age: 
\code{\link{age}}
}
\concept{Age}
\concept{Children}
\concept{Core}
\keyword{variable}
