% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_342.R
\name{agerborn}
\alias{agerborn}
\title{Age of r at born-again experience}
\source{
General Social Survey https://gss.norc.org
}
\description{
agerborn
}
\details{
Question How old were you when this occurred?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3065/vshow}{GSS Data Explorer page for \code{agerborn}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 25 \tab 29 \tab 30 \tab 31 \tab 32 \tab 36 \tab 39 \tab 52 \tab 60 \tab 68 \tab 75 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2756 \tab 1 \tab 5 \tab 4 \tab 4 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab 2812 \cr
   Total \tab 2756 \tab 1 \tab 5 \tab 4 \tab 4 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{chngrbrn}},
\code{\link{entity}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{numrborn}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}}
}
\concept{Spirituality and Religion}
\keyword{variable}
