% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{agewed}
\alias{agewed}
\title{Age when first married}
\source{
General Social Survey https://gss.norc.org
}
\description{
agewed
}
\details{
Question Are you currently--married, widowed, divorced, separated, or have you never been married?
A. IF EVER MARRIED:   How old were you when you first married?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/20/vshow}{GSS Data Explorer page for \code{agewed}}.
}

\seealso{
Other Core: 
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus10}},
\code{\link{marital}},
\code{\link{martype}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{prestg10}},
\code{\link{prestg105plus}}
}
\concept{Core}
\keyword{variable}
