% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{agewedcr}
\alias{agewedcr}
\title{R's age at current marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
agewedcr
}
\details{
Question How old were you when your (current) marriage began?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1481/vshow}{GSS Data Explorer page for \code{agewedcr}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 47 \tab 48 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 59 \tab 60 \tab 61 \tab 64 \tab 68 \tab 69 \tab 70 \tab 74 \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 2811 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 10 \tab 11 \tab 7 \tab 8 \tab 10 \tab 5 \tab 4 \tab 9 \tab 7 \tab 5 \tab 10 \tab 8 \tab 11 \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab 3 \tab 3 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab - \tab 2992 \cr
   Total \tab 2811 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 10 \tab 11 \tab 7 \tab 8 \tab 10 \tab 5 \tab 4 \tab 9 \tab 7 \tab 5 \tab 10 \tab 8 \tab 11 \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab 3 \tab 3 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{fstspyr}},
\code{\link{indfirst}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpicked}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{malive1}},
\code{\link{marnum}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{palive1}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind10}},
\code{\link{sbind80}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sbocc80}},
\code{\link{sboccindv}},
\code{\link{sboccstatus}},
\code{\link{sbocctag}},
\code{\link{sbpicked}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{spfam16}},
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre}},
\code{\link{sphedpre10}},
\code{\link{sphedpre105plus}},
\code{\link{sphedslf}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spsibs}},
\code{\link{spyrborn}},
\code{\link{wrkslfst}}

Other Family: 
\code{\link{adoptkid}},
\code{\link{agewed}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{balwkfam}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{chldben}},
\code{\link{chlddth}},
\code{\link{chldhrt}},
\code{\link{clsclsex}},
\code{\link{cousins}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce}},
\code{\link{engda}},
\code{\link{engson}},
\code{\link{evdiv}},
\code{\link{evkid}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{famlife}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famvswk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{feovrwrk}},
\code{\link{flextime}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{inlawnum}},
\code{\link{inperson}},
\code{\link{jobvsfam}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidcall}},
\code{\link{kidlived}},
\code{\link{kidmost}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{kidvisit}},
\code{\link{letters}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{madeath}},
\code{\link{malive}},
\code{\link{malive1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{marital}},
\code{\link{marunion}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{meetings}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{mykids}},
\code{\link{niecenep}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{numcntct}},
\code{\link{numprobs}},
\code{\link{obtohelp}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{othkids}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{padeath}},
\code{\link{palive}},
\code{\link{palive1}},
\code{\link{parleave}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{pawrkgrw}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{posslq}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relcall}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relmarry}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{rfamlook}},
\code{\link{sibcall}},
\code{\link{sibdeath}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibvisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{spdeath}},
\code{\link{spfalook}},
\code{\link{spfam16}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{sppart}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{widowed}},
\code{\link{wkvsfam}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}}

Other Marital Status: 
\code{\link{agewed}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{evdiv}},
\code{\link{hapmar}},
\code{\link{marital}},
\code{\link{marnum}},
\code{\link{marunion}},
\code{\link{posslqy}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{spdue}},
\code{\link{spmarnum}},
\code{\link{widowed}}
}
\concept{Family}
\concept{Family Mobility}
\concept{Marital Status}
\keyword{variable}
