% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_444.R
\name{aidhouse}
\alias{aidhouse}
\title{Gov. should provide housing to poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidhouse
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to . . .
I. Provide decent housing for those who can't afford  it.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should be
\item \verb{[2]} probably should be
\item \verb{[3]} probably should not be
\item \verb{[4]} definitely should not be
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4084/vshow}{GSS Data Explorer page for \code{aidhouse}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1990 \tab 155 \tab 236 \tab 588 \tab 231 \tab 66 \tab 57 \tab 39 \tab 1372 \cr
   1996 \tab 1572 \tab 240 \tab 572 \tab 295 \tab 106 \tab 61 \tab 58 \tab 2904 \cr
   2006 \tab 2992 \tab 494 \tab 647 \tab 280 \tab 70 \tab 26 \tab 1 \tab 4510 \cr
   2016 \tab 1477 \tab 375 \tab 654 \tab 251 \tab 66 \tab 31 \tab 13 \tab 2867 \cr
   Total \tab 6196 \tab 1345 \tab 2461 \tab 1057 \tab 308 \tab 175 \tab 111 \tab 11653 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidcol}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{equalize}},
\code{\link{hlthcare}},
\code{\link{protstrs}},
\code{\link{racists}},
\code{\link{revoltrs}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
