% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_444.R
\name{aidindus}
\alias{aidindus}
\title{Govts resp: assist industrial growth}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidindus
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to . . .
E. Provide industry with the help it needs to grow.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should be
\item \verb{[2]} probably should be
\item \verb{[3]} probably should not be
\item \verb{[4]} definitely should not be
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4080/vshow}{GSS Data Explorer page for \code{aidindus}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 107 \tab 289 \tab 159 \tab 72 \tab 39 \tab 11 \tab 1534 \cr
   1990 \tab 155 \tab 205 \tab 572 \tab 272 \tab 54 \tab 74 \tab 40 \tab 1372 \cr
   1996 \tab 1572 \tab 206 \tab 580 \tab 306 \tab 104 \tab 76 \tab 60 \tab 2904 \cr
   2006 \tab 2992 \tab 421 \tab 674 \tab 318 \tab 71 \tab 34 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 371 \tab 611 \tab 283 \tab 81 \tab 32 \tab 12 \tab 2867 \cr
   Total \tab 7053 \tab 1310 \tab 2726 \tab 1338 \tab 382 \tab 255 \tab 123 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{equalize}},
\code{\link{hlthcare}},
\code{\link{protstrs}},
\code{\link{racists}},
\code{\link{revoltrs}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
