% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{aidkids}
\alias{aidkids}
\title{Govts resp provide childcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidkids
}
\details{
Question On the whole, do you think it should or should not be the government''s responsibility to ...
Provide childcare for everyone who wants it?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should be
\item \verb{[2]} probably should be
\item \verb{[3]} probably should not be
\item \verb{[4]} definitely should not be
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4233/vshow}{GSS Data Explorer page for \code{aidkids}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 145 \tab 268 \tab 310 \tab 320 \tab 95 \tab 11 \tab 2765 \cr
   Total \tab 1616 \tab 145 \tab 268 \tab 310 \tab 320 \tab 95 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{comyear}},
\code{\link{demands}},
\code{\link{helpfrds}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
