% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{aidneedy}
\alias{aidneedy}
\title{Govt money for low income students}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidneedy
}
\details{
Question Some people think the government should provide financial assistance to college students. Others think the government should not provide such aid. In each of the circumstances listed below should the government provide grants that would not have to be paid back, provide loans which the student would have to pay back, or should the government not provide any financial assistance?
A. For students whose parents have a low income.
}
\section{Values}{

\itemize{
\item \code{1} govt give grants
\item \code{2} govt make loans
\item \code{3} no govt asst
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4025/vshow}{GSS Data Explorer page for \code{aidneedy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab govt give grants \tab govt make loans \tab no govt asst \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 279 \tab 351 \tab 32 \tab 6 \tab 9 \tab 1534 \cr
   Total \tab 857 \tab 279 \tab 351 \tab 32 \tab 6 \tab 9 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidsmart}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
