% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_551.R
\name{aidsads}
\alias{aidsads}
\title{Have govt info pgm on safe sex practices}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsads
}
\details{
Question Do you support or oppose the following measures to deal with AIDS?
B. Develop a government information program to promote safe sex practices, such as the use of condoms
}
\section{Values}{

\itemize{
\item \code{1} support
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5042/vshow}{GSS Data Explorer page for \code{aidsads}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab support \tab oppose \tab don't know \tab no answer \tab Total \cr
   1988 \tab 763 \tab 583 \tab 95 \tab 30 \tab 10 \tab 1481 \cr
   Total \tab 763 \tab 583 \tab 95 \tab 30 \tab 10 \tab 1481 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsmar}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{suiage3}},
\code{\link{suirac3}},
\code{\link{suireg3}}
}
\concept{Victims & AIDS}
\keyword{variable}
