% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_548.R
\name{aidsage2}
\alias{aidsage2}
\title{Age of aids victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsage2
}
\details{
Question We would like to know a few other things about that person.
C. How old (is/was) that person? (Is/Was) (he/she) . . . (READ LIST)
}
\section{Values}{

\itemize{
\item \verb{[1]} 10 years or under
\item \verb{[2]} 11-20
\item \verb{[3]} 21-40
\item \verb{[4]} 41 year or older
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5016/vshow}{GSS Data Explorer page for \code{aidsage2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 10 years or under \tab 11-20 \tab 21-40 \tab 41 year or older \tab no answer \tab Total \cr
   1989 \tab 1486 \tab 1 \tab 3 \tab 35 \tab 9 \tab 3 \tab 1537 \cr
   1990 \tab 1324 \tab 1 \tab 2 \tab 37 \tab 6 \tab 2 \tab 1372 \cr
   1991 \tab 1422 \tab 2 \tab 1 \tab 69 \tab 18 \tab 5 \tab 1517 \cr
   1993 \tab 1407 \tab 1 \tab 6 \tab 145 \tab 39 \tab 8 \tab 1606 \cr
   Total \tab 5639 \tab 5 \tab 12 \tab 286 \tab 72 \tab 18 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsrac2}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidswho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
