% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsage3}
\alias{aidsage3}
\title{Age of aids victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsage3
}
\details{
Question We would like to know a few other things about that person.
C. How old (is/was) that person? (Is/Was) (he/she) . . . (READ LIST)
}
\section{Values}{

\itemize{
\item \verb{[1]} 10 years or under
\item \verb{[2]} 11-20
\item \verb{[3]} 21-40
\item \verb{[4]} 41 year or older
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5022/vshow}{GSS Data Explorer page for \code{aidsage3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 21-40 \tab 41 year or older \tab no answer \tab 11-20 \tab 10 years or under \tab not available in this year \tab Total \cr
   1989 \tab 1517 \tab 15 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1347 \tab 19 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1466 \tab 37 \tab 5 \tab 5 \tab 3 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 1490 \tab 88 \tab 19 \tab 6 \tab 2 \tab 1 \tab - \tab 1606 \cr
   Total \tab 5820 \tab 159 \tab 30 \tab 15 \tab 6 \tab 2 \tab 0 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssch}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}},
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suiage3}},
\code{\link{suiknew}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Aids: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidscndm}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidslook}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{hivkiss}},
\code{\link{hivvac}}
}
\concept{Aids}
\concept{Victims & AIDS}
\keyword{variable}
