% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{aidscndm}
\alias{aidscndm}
\title{Condom can reduce aids}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidscndm
}
\details{
Question B. Can people reduce their chance of getting the AIDS virus by using a condom every time they have sex?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5078/vshow}{GSS Data Explorer page for \code{aidscndm}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2014 \tab 707 \tab 1663 \tab 95 \tab 39 \tab 34 \tab 2538 \cr
   Total \tab 707 \tab 1663 \tab 95 \tab 39 \tab 34 \tab 2538 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{hivkiss}},
\code{\link{hivvac}},
\code{\link{sexornt}}
}
\concept{High Risk Behaviors}
\keyword{variable}
