% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_548.R
\name{aidsdead}
\alias{aidsdead}
\title{Has aids victim closest to r died}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsdead
}
\details{
Question We would like to know a few other things about that person.
A. Is that person currently living, or has that person died?
}
\section{Values}{

\itemize{
\item \verb{[1]} living
\item \verb{[2]} died
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5008/vshow}{GSS Data Explorer page for \code{aidsdead}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab living \tab died \tab no answer \tab don't know \tab Total \cr
   1988 \tab 1342 \tab 49 \tab 86 \tab 4 \tab - \tab 1481 \cr
   1989 \tab 1392 \tab 44 \tab 98 \tab 2 \tab 1 \tab 1537 \cr
   1990 \tab 1195 \tab 52 \tab 118 \tab 2 \tab 5 \tab 1372 \cr
   1991 \tab 1275 \tab 64 \tab 164 \tab 6 \tab 8 \tab 1517 \cr
   1993 \tab 1214 \tab 91 \tab 287 \tab 7 \tab 7 \tab 1606 \cr
   Total \tab 6418 \tab 300 \tab 753 \tab 21 \tab 21 \tab 7513 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsded2}},
\code{\link{aidsrac2}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidswho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
