% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{aidsfare}
\alias{aidsfare}
\title{Pay aids victims disability benefits}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsfare
}
\details{
Question Do you support or oppose the following measures to deal with AIDS?
D. Make victims with AIDS eligible for disability benefits
}
\section{Values}{

\itemize{
\item \code{1} support
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5048/vshow}{GSS Data Explorer page for \code{aidsfare}}.
}

\seealso{
Other Split Ballots: 
\code{\link{partners}},
\code{\link{sexfreq}}
}
\concept{Aids}
\concept{Split Ballots}
\concept{Victims & AIDS}
\keyword{variable}
