% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_551.R
\name{aidsmar}
\alias{aidsmar}
\title{Have testing for aids before marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsmar
}
\details{
Question Do you support or oppose the following measures to deal with AIDS?
A. Conduct mandatory testing for the AIDS virus before marriage
}
\section{Values}{

\itemize{
\item \verb{[1]} support
\item \verb{[2]} oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5045/vshow}{GSS Data Explorer page for \code{aidsmar}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab support \tab oppose \tab don't know \tab no answer \tab Total \cr
   1988 \tab 718 \tab 647 \tab 80 \tab 26 \tab 10 \tab 1481 \cr
   Total \tab 718 \tab 647 \tab 80 \tab 26 \tab 10 \tab 1481 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{suiage3}},
\code{\link{suirac3}},
\code{\link{suireg3}}
}
\concept{Victims & AIDS}
\keyword{variable}
