% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{aidsmart}
\alias{aidsmart}
\title{Govt money for high achieving students}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsmart
}
\details{
Question Some people think the government should provide financial assistance to college students. Others think the government should not provide such aid. In each of the circumstances listed below should the government provide grants that would not have to be paid back, provide loans which the student would have to pay back, or should the government not provide any financial assistance?
B. For students who have outstanding grades in high school.
}
\section{Values}{

\itemize{
\item \verb{[1]} govt give grants
\item \verb{[2]} govt make loans
\item \verb{[3]} no govt asst
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4026/vshow}{GSS Data Explorer page for \code{aidsmart}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab govt give grants \tab govt make loans \tab no govt asst \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 226 \tab 364 \tab 57 \tab 13 \tab 17 \tab 1534 \cr
   Total \tab 857 \tab 226 \tab 364 \tab 57 \tab 13 \tab 17 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
