% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_548.R
\name{aidsreg}
\alias{aidsreg}
\title{Region of aids victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsreg
}
\details{
Question We would like to know a few other things about that person.
E. What state did this person live in?
}
\section{Values}{

\itemize{
\item \verb{[0]} foreign
\item \verb{[1]} new england
\item \verb{[2]} middle atlantic
\item \verb{[3]} e. nor. central
\item \verb{[4]} w. nor. central
\item \verb{[5]} south atlantic
\item \verb{[6]} e. sou. central
\item \verb{[7]} w. sou. central
\item \verb{[8]} mountain
\item \verb{[9]} pacific
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5012/vshow}{GSS Data Explorer page for \code{aidsreg}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab new england \tab middle atlantic \tab e. nor. central \tab w. nor. central \tab south atlantic \tab e. sou. central \tab w. sou. central \tab mountain \tab pacific \tab no answer \tab foreign \tab not available in this year \tab Total \cr
   1989 \tab 1392 \tab 4 \tab 24 \tab 14 \tab 11 \tab 27 \tab 3 \tab 17 \tab 8 \tab 34 \tab 3 \tab - \tab - \tab 1537 \cr
   1990 \tab 1195 \tab 9 \tab 39 \tab 22 \tab 11 \tab 32 \tab 6 \tab 10 \tab 10 \tab 34 \tab 2 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 1275 \tab 12 \tab 39 \tab 30 \tab 14 \tab 39 \tab 5 \tab 18 \tab 12 \tab 61 \tab 9 \tab 3 \tab - \tab 1517 \cr
   1993 \tab 1214 \tab 22 \tab 64 \tab 44 \tab 17 \tab 76 \tab 16 \tab 31 \tab 15 \tab 93 \tab 9 \tab 5 \tab - \tab 1606 \cr
   Total \tab 5076 \tab 47 \tab 166 \tab 110 \tab 53 \tab 174 \tab 30 \tab 76 \tab 45 \tab 222 \tab 23 \tab 10 \tab 0 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssch}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}},
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suiage3}},
\code{\link{suiknew}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Aids: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidscndm}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidslook}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{hivkiss}},
\code{\link{hivvac}}
}
\concept{Aids}
\concept{Victims & AIDS}
\keyword{variable}
