% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsreg2}
\alias{aidsreg2}
\title{Region of aids victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsreg2
}
\details{
Question We would like to know a few other things about that person.
E. What state did this person live in?
}
\section{Values}{

\itemize{
\item \code{0} foreign
\item \code{1} new england
\item \code{2} middle atlantic
\item \code{3} e. nor. central
\item \code{4} w. nor. central
\item \code{5} south atlantic
\item \code{6} e. sou. central
\item \code{7} w. sou. central
\item \code{8} mountain
\item \code{9} pacific
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5018/vshow}{GSS Data Explorer page for \code{aidsreg2}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab new england \tab middle atlantic \tab e. nor. central \tab w. nor. central \tab south atlantic \tab e. sou. central \tab w. sou. central \tab pacific \tab no answer \tab mountain \tab foreign \tab Total \cr
   1989 \tab 1486 \tab 2 \tab 10 \tab 3 \tab 1 \tab 10 \tab 1 \tab 9 \tab 10 \tab 5 \tab - \tab - \tab 1537 \cr
   1990 \tab 1324 \tab - \tab 7 \tab 6 \tab 2 \tab 13 \tab 1 \tab 4 \tab 9 \tab 2 \tab 4 \tab - \tab 1372 \cr
   1991 \tab 1422 \tab 3 \tab 16 \tab 11 \tab 3 \tab 18 \tab 1 \tab 7 \tab 29 \tab 5 \tab 2 \tab - \tab 1517 \cr
   1993 \tab 1407 \tab 18 \tab 33 \tab 23 \tab 2 \tab 36 \tab 8 \tab 17 \tab 40 \tab 10 \tab 9 \tab 3 \tab 1606 \cr
   Total \tab 5639 \tab 23 \tab 66 \tab 43 \tab 8 \tab 77 \tab 11 \tab 37 \tab 88 \tab 22 \tab 15 \tab 3 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
