% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsreg3}
\alias{aidsreg3}
\title{Region of aids victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsreg3
}
\details{
Question We would like to know a few other things about that person.
E. What state did this person live in?
}
\section{Values}{

\itemize{
\item \verb{[0]} foreign
\item \verb{[1]} new england
\item \verb{[2]} middle atlantic
\item \verb{[3]} e. nor. central
\item \verb{[4]} w. nor. central
\item \verb{[5]} south atlantic
\item \verb{[6]} e. sou. central
\item \verb{[7]} w. sou. central
\item \verb{[8]} mountain
\item \verb{[9]} pacific
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5024/vshow}{GSS Data Explorer page for \code{aidsreg3}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab new england \tab middle atlantic \tab e. nor. central \tab south atlantic \tab w. sou. central \tab pacific \tab no answer \tab e. sou. central \tab mountain \tab w. nor. central \tab Total \cr
   1989 \tab 1517 \tab 1 \tab 3 \tab 1 \tab 4 \tab 4 \tab 3 \tab 4 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1347 \tab - \tab 7 \tab 4 \tab 5 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 1466 \tab 1 \tab 8 \tab 6 \tab 8 \tab 2 \tab 16 \tab 7 \tab - \tab 1 \tab 2 \tab 1517 \cr
   1993 \tab 1490 \tab 5 \tab 24 \tab 12 \tab 21 \tab 9 \tab 25 \tab 6 \tab 3 \tab 8 \tab 3 \tab 1606 \cr
   Total \tab 5820 \tab 7 \tab 42 \tab 23 \tab 38 \tab 16 \tab 49 \tab 18 \tab 4 \tab 10 \tab 5 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
