% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_551.R
\name{aidssch}
\alias{aidssch}
\title{Prohibit aids-students at public schools}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidssch
}
\details{
Question Do you support or oppose the following measures to deal with AIDS?
A. Prohibit students with the AIDS virus from attending public school
}
\section{Values}{

\itemize{
\item \code{1} support
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5041/vshow}{GSS Data Explorer page for \code{aidssch}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsmar}},
\code{\link{aidssxed}},
\code{\link{suiage3}},
\code{\link{suirac3}},
\code{\link{suireg3}}
}
\concept{Victims & AIDS}
\keyword{variable}
