% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_548.R
\name{aidssex2}
\alias{aidssex2}
\title{Sex of aids victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidssex2
}
\details{
Question We would like to know a few other things about that person.
B. (Is/Was) that person male or female?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5015/vshow}{GSS Data Explorer page for \code{aidssex2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1989 \tab 1486 \tab 41 \tab 6 \tab 4 \tab 1537 \cr
   1990 \tab 1324 \tab 46 \tab - \tab 2 \tab 1372 \cr
   1991 \tab 1422 \tab 81 \tab 10 \tab 4 \tab 1517 \cr
   1993 \tab 1407 \tab 175 \tab 17 \tab 7 \tab 1606 \cr
   Total \tab 5639 \tab 343 \tab 33 \tab 17 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsrac2}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidssex}},
\code{\link{aidswho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
