% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidssex3}
\alias{aidssex3}
\title{Sex of aids victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidssex3
}
\details{
Question We would like to know a few other things about that person.
B. (Is/Was) that person male or female?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5021/vshow}{GSS Data Explorer page for \code{aidssex3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1989 \tab 1517 \tab 16 \tab 1 \tab 3 \tab 1537 \cr
   1990 \tab 1347 \tab 23 \tab 1 \tab 1 \tab 1372 \cr
   1991 \tab 1466 \tab 43 \tab 3 \tab 5 \tab 1517 \cr
   1993 \tab 1490 \tab 101 \tab 9 \tab 6 \tab 1606 \cr
   Total \tab 5820 \tab 183 \tab 14 \tab 15 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
