% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{aidswho}
\alias{aidswho}
\title{R's relationship to closest aids victim}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidswho
}
\details{
Question Think about the person you have known best, living or dead, who came down with AIDS. Please tell me the letter of the category on the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \verb{[1]} husband or wife
\item \verb{[2]} partner or lover
\item \verb{[3]} son or daughter
\item \verb{[4]} other relative
\item \verb{[5]} friend
\item \verb{[6]} neighbor
\item \verb{[7]} co-worker
\item \verb{[8]} acquaintance
\item \verb{[9]} patient
\item \verb{[10]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1993 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5007/vshow}{GSS Data Explorer page for \code{aidswho}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab husband or wife \tab other relative \tab friend \tab neighbor \tab co-worker \tab acquaintance \tab patient \tab other \tab no answer \tab partner or lover \tab son or daughter \tab not available in this year \tab Total \cr
   1988 \tab 1342 \tab 1 \tab 5 \tab 37 \tab 7 \tab 21 \tab 52 \tab 6 \tab 7 \tab 3 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1392 \tab - \tab 19 \tab 40 \tab 2 \tab 12 \tab 53 \tab 3 \tab 11 \tab 3 \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 1195 \tab - \tab 23 \tab 41 \tab 5 \tab 22 \tab 66 \tab 6 \tab 11 \tab 1 \tab 1 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 1348 \tab 2 \tab 22 \tab 48 \tab 7 \tab 15 \tab 52 \tab 8 \tab 12 \tab 1 \tab 2 \tab - \tab - \tab 1517 \cr
   1993 \tab 1214 \tab 1 \tab 52 \tab 125 \tab 19 \tab 47 \tab 112 \tab 9 \tab 16 \tab 7 \tab 2 \tab 2 \tab - \tab 1606 \cr
   Total \tab 6491 \tab 4 \tab 121 \tab 291 \tab 40 \tab 117 \tab 335 \tab 32 \tab 57 \tab 15 \tab 7 \tab 3 \tab 0 \tab 7513 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssch}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}},
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suiage3}},
\code{\link{suiknew}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Aids: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidscndm}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidslook}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{hivkiss}},
\code{\link{hivvac}}
}
\concept{Aids}
\concept{Victims & AIDS}
\keyword{variable}
