% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{airpollu}
\alias{airpollu}
\title{Neighborhood affected by air pollution over the last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
airpollu
}
\details{
Question To what degree was your neighborhood affected by air polution over the last twelve months?
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[2]} to a small extent
\item \verb{[3]} to some extent
\item \verb{[4]} to a great extent
\item \verb{[5]} to a very great extent
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7501/vshow}{GSS Data Explorer page for \code{airpollu}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab not at all \tab to a small extent \tab to some extent \tab to a great extent \tab to a very great extent \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 661 \tab 563 \tab 348 \tab 96 \tab 49 \tab 105 \tab 200 \tab 7 \tab 4032 \cr
   Total \tab 2003 \tab 661 \tab 563 \tab 348 \tab 96 \tab 49 \tab 105 \tab 200 \tab 7 \tab 4032 \cr
}
}

\keyword{variable}
