% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_49.R
\name{alienat1}
\alias{alienat1}
\title{People running country don't care}
\source{
General Social Survey https://gss.norc.org
}
\description{
alienat1
}
\details{
Question Now I want to read you some things some people have told us they have felt from time to time. Do you tend to feel or not . . . (READ LIST)
A. The people running the country don''t really care what happens to you.
}
\section{Values}{

\itemize{
\item \verb{[1]} feel
\item \verb{[2]} not feel
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/428/vshow}{GSS Data Explorer page for \code{alienat1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab feel \tab not feel \tab don't know \tab no answer \tab Total \cr
   1978 \tab - \tab 779 \tab 678 \tab 63 \tab 12 \tab 1532 \cr
   Total \tab 0 \tab 779 \tab 678 \tab 63 \tab 12 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{hapmar}},
\code{\link{happy}}
}
\concept{Core}
\keyword{variable}
