% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_49.R
\name{alienat2}
\alias{alienat2}
\title{Rich get richer, poor poorer}
\source{
General Social Survey https://gss.norc.org
}
\description{
alienat2
}
\details{
Question Now I want to read you some things some people have told us they have felt from time to time. Do you tend to feel or not . . . (READ LIST)
B. The rich get richer and the poor get poorer.
}
\section{Values}{

\itemize{
\item \verb{[1]} feel
\item \verb{[2]} not feel
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/429/vshow}{GSS Data Explorer page for \code{alienat2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab feel \tab not feel \tab don't know \tab no answer \tab Total \cr
   1978 \tab - \tab 1120 \tab 355 \tab 43 \tab 14 \tab 1532 \cr
   Total \tab 0 \tab 1120 \tab 355 \tab 43 \tab 14 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{alienat1}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{hapmar}},
\code{\link{happy}}
}
\concept{Core}
\keyword{variable}
