% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_49.R
\name{alienat5}
\alias{alienat5}
\title{Powerful people take advantage of you}
\source{
General Social Survey https://gss.norc.org
}
\description{
alienat5
}
\details{
Question Now I want to read you some things some people have told us they have felt from time to time. Do you tend to feel or not . . . (READ LIST)
E. Most people with power try to take advantage of people like yourself.
}
\section{Values}{

\itemize{
\item \code{1} feel
\item \code{2} not feel
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/432/vshow}{GSS Data Explorer page for \code{alienat5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab feel \tab not feel \tab don't know \tab no answer \tab Total \cr
   1978 \tab - \tab 845 \tab 625 \tab 50 \tab 12 \tab 1532 \cr
   Total \tab 0 \tab 845 \tab 625 \tab 50 \tab 12 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat6}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{hapmar}},
\code{\link{happy}}
}
\concept{Core}
\keyword{variable}
