% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{alike1}
\alias{alike1}
\title{How alike: orange & banana}
\source{
General Social Survey https://gss.norc.org
}
\description{
alike1
}
\details{
Question Now I'd like to turn to something a bit different. This section is about reasoning abilities, that is, about how people think. Most people take tests of their reasoning abilities and there is a lot of interest in how these abilities may change over time. What I'd like you to do is tell me how two things are alike.
A. In what way are an orange and a banana alike?
}
\section{Values}{

\itemize{
\item \code{0} completely wrong
\item \code{1} partially right
\item \code{2} completely right
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1467/vshow}{GSS Data Explorer page for \code{alike1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab completely wrong \tab partially right \tab completely right \tab don't know \tab no answer \tab Total \cr
   1994 \tab - \tab 179 \tab 354 \tab 2410 \tab 31 \tab 18 \tab 2992 \cr
   Total \tab 0 \tab 179 \tab 354 \tab 2410 \tab 31 \tab 18 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike2}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{indfirst}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{wrkslfst}}
}
\concept{Family Mobility}
\keyword{variable}
