% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{alike2}
\alias{alike2}
\title{How alike: dog & lion}
\source{
General Social Survey https://gss.norc.org
}
\description{
alike2
}
\details{
Question Now I'd like to turn to something a bit different. This section is about reasoning abilities, that is, about how people think. Most people take tests of their reasoning abilities and there is a lot of interest in how these abilities may change over time. What I'd like you to do is tell me how two things are alike.
B. In what way are a dog and a lion alike?
}
\section{Values}{

\itemize{
\item \code{0} completely wrong
\item \code{1} partially right
\item \code{2} completely right
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1468/vshow}{GSS Data Explorer page for \code{alike2}}.
}

\seealso{
Other Family Mobility: 
\code{\link{alike1}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{indfirst}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{wrkslfst}}

Other Knowledge: 
\code{\link{alike1}}

Other Reasoning Abilities: 
\code{\link{alike1}}
}
\concept{Family Mobility}
\concept{Knowledge}
\concept{Reasoning Abilities}
\keyword{variable}
