% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_156.R
\name{alike4}
\alias{alike4}
\title{How alike: egg & seed}
\source{
General Social Survey https://gss.norc.org
}
\description{
alike4
}
\details{
Question Now I'd like to turn to something a bit different. This section is about reasoning abilities, that is, about how people think. Most people take tests of their reasoning abilities and there is a lot of interest in how these abilities may change over time. What I'd like you to do is tell me how two things are alike.
D. In what way are an egg and a seed alike?
}
\section{Values}{

\itemize{
\item \verb{[0]} completely wrong
\item \verb{[1]} partially right
\item \verb{[2]} completely right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1470/vshow}{GSS Data Explorer page for \code{alike4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab completely wrong \tab partially right \tab completely right \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab - \tab 828 \tab 1107 \tab 820 \tab 208 \tab 29 \tab - \tab 2992 \cr
   Total \tab 0 \tab 828 \tab 1107 \tab 820 \tab 208 \tab 29 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{fstspyr}},
\code{\link{indfirst}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpicked}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{malive1}},
\code{\link{marnum}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{palive1}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind10}},
\code{\link{sbind80}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sbocc80}},
\code{\link{sboccindv}},
\code{\link{sboccstatus}},
\code{\link{sbocctag}},
\code{\link{sbpicked}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{spfam16}},
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre}},
\code{\link{sphedpre10}},
\code{\link{sphedpre105plus}},
\code{\link{sphedslf}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spsibs}},
\code{\link{spyrborn}},
\code{\link{wrkslfst}}

Other Knowledge: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{anheat}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{boyorgrl}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{electron}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{governor}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{h2olife}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{nanoben}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{poleff20}},
\code{\link{radioact}},
\code{\link{schlhead}},
\code{\link{scifrom}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{seeksci}},
\code{\link{seesand}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{upbreath}},
\code{\link{usrep}},
\code{\link{viruses}},
\code{\link{weighing}},
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Reasoning Abilities: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}}
}
\concept{Family Mobility}
\concept{Knowledge}
\concept{Reasoning Abilities}
\keyword{variable}
