% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_156.R
\name{alike7}
\alias{alike7}
\title{How alike: fly & tree}
\source{
General Social Survey https://gss.norc.org
}
\description{
alike7
}
\details{
Question Now I'd like to turn to something a bit different. This section is about reasoning abilities, that is, about how people think. Most people take tests of their reasoning abilities and there is a lot of interest in how these abilities may change over time. What I'd like you to do is tell me how two things are alike.
G. In what way are a fly and a tree alike?
}
\section{Values}{

\itemize{
\item \verb{[0]} completely wrong
\item \verb{[1]} partially right
\item \verb{[2]} completely right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1473/vshow}{GSS Data Explorer page for \code{alike7}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab completely wrong \tab partially right \tab completely right \tab don't know \tab no answer \tab Total \cr
   1994 \tab - \tab 1410 \tab 387 \tab 631 \tab 530 \tab 34 \tab 2992 \cr
   Total \tab 0 \tab 1410 \tab 387 \tab 631 \tab 530 \tab 34 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike8}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspyr}},
\code{\link{marnum}}
}
\concept{Family Mobility}
\keyword{variable}
