% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{allben}
\alias{allben}
\title{Everyone benefit from traditional family}
\source{
General Social Survey https://gss.norc.org
}
\description{
allben
}
\details{
Question In many married couples, women take the main responsibility for the care of the home and children, while men take the main responsibility for supporting the family financially. Who benefits from this: men, women, children, everyone, or no one? PROBE: Anyone else? UNTIL R SAYS NO.
D. Everyone
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1681/vshow}{GSS Data Explorer page for \code{allben}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 837 \tab 397 \tab 77 \tab 149 \tab 2904 \cr
   Total \tab 1444 \tab 837 \tab 397 \tab 77 \tab 149 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{chldben}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{parleave}},
\code{\link{womenben}}
}
\concept{Gender}
\keyword{variable}
