% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{alloflfe}
\alias{alloflfe}
\title{Experience connection to all of life}
\source{
General Social Survey https://gss.norc.org
}
\description{
alloflfe
}
\details{
Question Experience a connection to all of life
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never or almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3083/vshow}{GSS Data Explorer page for \code{alloflfe}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never or almost never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 173 \tab 368 \tab 285 \tab 182 \tab 161 \tab 103 \tab 49 \tab 19 \tab 2812 \cr
   Total \tab 1472 \tab 173 \tab 368 \tab 285 \tab 182 \tab 161 \tab 103 \tab 49 \tab 19 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}}
}
\concept{Spirituality and Religion}
\keyword{variable}
