% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{allorglevel}
\alias{allorglevel}
\title{All levels of org are involved in stress prevention}
\source{
General Social Survey https://gss.norc.org
}
\description{
allorglevel
}
\details{
Question (To what extent to do you agree or disagree with the following statements?) In my organization, the prevention of stress involves all levels of the organization.
}
\section{Values}{

\itemize{
\item \code{1} strongly disagree
\item \code{2} disagree
\item \code{3} neither agree nor disagree
\item \code{4} agree
\item \code{5} strongly agree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7858/vshow}{GSS Data Explorer page for \code{allorglevel}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly disagree \tab disagree \tab neither agree nor disagree \tab agree \tab strongly agree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1549 \tab 198 \tab 300 \tab 582 \tab 674 \tab 156 \tab 36 \tab 23 \tab 26 \tab 3544 \cr
   Total \tab 1549 \tab 198 \tab 300 \tab 582 \tab 674 \tab 156 \tab 36 \tab 23 \tab 26 \tab 3544 \cr
}
}

\keyword{variable}
