% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_545.R
\name{allowcom}
\alias{allowcom}
\title{Allow communist to speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
allowcom
}
\details{
Question D. Now, I would like to ask you some questions about a man who admits he is a Communist. Suppose this admitted Communist wanted to make a speech in your community. Should he be allowed to speak?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, allow
\item \verb{[2]} no, not allow
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4987/vshow}{GSS Data Explorer page for \code{allowcom}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, allow \tab no, not allow \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1989 \tab 1029 \tab 303 \tab 161 \tab 13 \tab 31 \tab - \tab 1537 \cr
   Total \tab 1029 \tab 303 \tab 161 \tab 13 \tab 31 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Civil Liberties: 
\code{\link{allowmar}},
\code{\link{allowrac}},
\code{\link{antirel}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdcom}},
\code{\link{forbdmar}},
\code{\link{forbdrac}},
\code{\link{irrelart}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{protstrs}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{racists}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{racteach}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{revmeet}},
\code{\link{revoltrs}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{wirtap}}

Other Civil Liberties: 
\code{\link{allowmar}},
\code{\link{allowrac}},
\code{\link{antirel}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdcom}},
\code{\link{forbdmar}},
\code{\link{forbdrac}},
\code{\link{irrelart}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{protstrs}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{racists}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{racteach}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{revmeet}},
\code{\link{revoltrs}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{wirtap}}
}
\concept{Civil Liberties}
\keyword{variable}
