% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{ambornin}
\alias{ambornin}
\title{How important to have been born in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
ambornin
}
\details{
Question Some people say the following things are important for being truly American. Others say they are not important. How important do you think each of the following is...
A. To have been born in America
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} fairly important
\item \verb{[3]} not very important
\item \verb{[4]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4827/vshow}{GSS Data Explorer page for \code{ambornin}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 546 \tab 365 \tab 250 \tab 164 \tab 26 \tab 16 \tab 2904 \cr
   2004 \tab 1596 \tab 679 \tab 245 \tab 196 \tab 74 \tab 21 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 530 \tab 283 \tab 292 \tab 137 \tab 30 \tab 2 \tab 2538 \cr
   Total \tab 4397 \tab 1755 \tab 893 \tab 738 \tab 375 \tab 77 \tab 19 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{onenatn}}
}
\concept{ISSP National Identity}
\keyword{variable}
