% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amcit}
\alias{amcit}
\title{How important to have american citizenship}
\source{
General Social Survey https://gss.norc.org
}
\description{
amcit
}
\details{
Question Some people say the following things are important for being truly American. Others say they are not important. How important do you think each of the following is...
B. To have American citizenship
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} fairly important
\item \verb{[3]} not very important
\item \verb{[4]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4828/vshow}{GSS Data Explorer page for \code{amcit}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 1007 \tab 232 \tab 68 \tab 32 \tab 10 \tab 18 \tab 2904 \cr
   2004 \tab 1596 \tab 999 \tab 159 \tab 37 \tab 11 \tab 9 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 903 \tab 272 \tab 56 \tab 28 \tab 12 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 2909 \tab 663 \tab 161 \tab 71 \tab 31 \tab 22 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcitizn}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
