% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amcitizn}
\alias{amcitizn}
\title{Agree i would rather be a citizen of america}
\source{
General Social Survey https://gss.norc.org
}
\description{
amcitizn
}
\details{
Question How much do you agree or disagree with the following statements?
A. I would rather be a citizen of America than of any other country in the world
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4834/vshow}{GSS Data Explorer page for \code{amcitizn}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}

Other United States: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
