% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{amcult}
\alias{amcult}
\title{It is impossible to become fully american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amcult
}
\details{
Question How much do you agree or disagree with the following statements?
A. It is impossible for people who do not share American customs and traditions to become fully American
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4856/vshow}{GSS Data Explorer page for \code{amcult}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
