% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amenglsh}
\alias{amenglsh}
\title{How important to be able to speak english}
\source{
General Social Survey https://gss.norc.org
}
\description{
amenglsh
}
\details{
Question Some people say the following things are important for being truly American. Others say they are not important. How important do you think each of the following is...
D. To be able to speak English
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} fairly important
\item \code{3} not very important
\item \code{4} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4830/vshow}{GSS Data Explorer page for \code{amenglsh}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 956 \tab 290 \tab 69 \tab 25 \tab 11 \tab 16 \tab 2904 \cr
   2004 \tab 1596 \tab 1011 \tab 166 \tab 28 \tab 8 \tab 2 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 918 \tab 267 \tab 53 \tab 26 \tab 7 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 2885 \tab 723 \tab 150 \tab 59 \tab 20 \tab 20 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
