% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amfeel}
\alias{amfeel}
\title{How important to feel american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amfeel
}
\details{
Question Some people say the following things are important for being truly American. Others say they are not important. How important do you think each of the following is...
G. To feel American
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} fairly important
\item \verb{[3]} not very important
\item \verb{[4]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4833/vshow}{GSS Data Explorer page for \code{amfeel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 821 \tab 331 \tab 130 \tab 46 \tab 22 \tab 17 \tab 2904 \cr
   2004 \tab 1596 \tab 812 \tab 284 \tab 81 \tab 14 \tab 24 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 721 \tab 329 \tab 125 \tab 56 \tab 40 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 2354 \tab 944 \tab 336 \tab 116 \tab 86 \tab 21 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amenglsh}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
