% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_54.R
\name{amicable}
\alias{amicable}
\title{Gets along well with other children}
\source{
General Social Survey https://gss.norc.org
}
\description{
amicable
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
8. That he gets along well with other children
}
\section{Values}{

\itemize{
\item \code{1} 1 most desirable
\item \code{2} 3 most desirable
\item \code{3} not mentioned
\item \code{4} 3 least desirable
\item \code{5} 1 least desirable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1975, 1976, 1978, 1980, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/481/vshow}{GSS Data Explorer page for \code{amicable}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 30 \tab 214 \tab 1021 \tab 180 \tab 55 \tab 4 \tab - \tab 1504 \cr
   1975 \tab - \tab 34 \tab 171 \tab 1051 \tab 182 \tab 42 \tab 10 \tab - \tab 1490 \cr
   1976 \tab - \tab 21 \tab 180 \tab 1076 \tab 171 \tab 41 \tab 10 \tab - \tab 1499 \cr
   1978 \tab - \tab 28 \tab 141 \tab 1128 \tab 165 \tab 58 \tab 10 \tab 2 \tab 1532 \cr
   1980 \tab 975 \tab 11 \tab 59 \tab 338 \tab 69 \tab 12 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 26 \tab 139 \tab 1195 \tab 175 \tab 44 \tab 20 \tab - \tab 1599 \cr
   Total \tab 975 \tab 150 \tab 904 \tab 5809 \tab 942 \tab 252 \tab 57 \tab 3 \tab 9092 \cr
}
}

\seealso{
Other Core: 
\code{\link{clean}},
\code{\link{consider}},
\code{\link{control}},
\code{\link{honest}},
\code{\link{judgment}},
\code{\link{obeys}},
\code{\link{responsi}},
\code{\link{role}},
\code{\link{success}}
}
\concept{Core}
\keyword{variable}
