% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_57.R
\name{amicablz}
\alias{amicablz}
\title{Gets along with other children-version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
amicablz
}
\details{
Question Please look at the qualities listed on this card. (SHOW CARD:
LET R. INSPECT) All of the qualities may be desirable for a child to have, but could you tell me whether the quality is extremely important, very important, fairly important, not too important, or not at all important. READ AND RECORD RESPONSES FOR EACH.
H. That he gets along well with other children
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely impt
\item \verb{[2]} very impt
\item \verb{[3]} fairly impt
\item \verb{[4]} not too impt
\item \verb{[5]} not at all impt
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/507/vshow}{GSS Data Explorer page for \code{amicablz}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab extremely impt \tab very impt \tab fairly impt \tab not too impt \tab don't know \tab no answer \tab Total \cr
   1980 \tab 985 \tab 101 \tab 246 \tab 125 \tab 6 \tab 1 \tab 4 \tab 1468 \cr
   Total \tab 985 \tab 101 \tab 246 \tab 125 \tab 6 \tab 1 \tab 4 \tab 1468 \cr
}
}

\seealso{
Other Core: 
\code{\link{considez}},
\code{\link{controlz}},
\code{\link{interesz}},
\code{\link{obey}},
\code{\link{obeysz}},
\code{\link{popular}},
\code{\link{responsz}},
\code{\link{rolez}},
\code{\link{studiouz}}
}
\concept{Core}
\keyword{variable}
