% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{amimp}
\alias{amimp}
\title{How important is being an american to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
amimp
}
\details{
Question How important is being an AMERICAN to you, where 0 is not at all important and 10 is the most important thing In your life
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[10]} most
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1787/vshow}{GSS Data Explorer page for \code{amimp}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab not at all \tab 1 \tab most \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2199 \tab 11 \tab 5 \tab 315 \tab 7 \tab 8 \tab 13 \tab 67 \tab 31 \tab 67 \tab 98 \tab 68 \tab 5 \tab 10 \tab 2904 \cr
   Total \tab 2199 \tab 11 \tab 5 \tab 315 \tab 7 \tab 8 \tab 13 \tab 67 \tab 31 \tab 67 \tab 98 \tab 68 \tab 5 \tab 10 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
