% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amlived}
\alias{amlived}
\title{How important to have lived in america for life}
\source{
General Social Survey https://gss.norc.org
}
\description{
amlived
}
\details{
Question Some people say the following things are important for being truly American. Others say they are not important. How important do you think each of the following is...
C. To have lived in America for most of one's life
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} fairly important
\item \verb{[3]} not very important
\item \verb{[4]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4829/vshow}{GSS Data Explorer page for \code{amlived}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 587 \tab 381 \tab 270 \tab 86 \tab 20 \tab 23 \tab 2904 \cr
   2004 \tab 1596 \tab 706 \tab 271 \tab 200 \tab 28 \tab 10 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 524 \tab 358 \tab 278 \tab 90 \tab 21 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 1817 \tab 1010 \tab 748 \tab 204 \tab 51 \tab 27 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
