% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{amownway}
\alias{amownway}
\title{America should follow its own interests}
\source{
General Social Survey https://gss.norc.org
}
\description{
amownway
}
\details{
Question How much do you agree or disagree with the following statements?
D. America should follow its own interests, even if this leads to conflicts with other nations
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4853/vshow}{GSS Data Explorer page for \code{amownway}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 154 \tab 418 \tab 351 \tab 329 \tab 38 \tab 51 \tab 26 \tab 2904 \cr
   2004 \tab 1596 \tab 152 \tab 402 \tab 291 \tab 294 \tab 40 \tab 36 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 114 \tab 440 \tab 276 \tab 315 \tab 57 \tab 67 \tab 5 \tab 2538 \cr
   Total \tab 4397 \tab 420 \tab 1260 \tab 918 \tab 938 \tab 135 \tab 154 \tab 32 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
