% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{amproud1}
\alias{amproud1}
\title{How proud being american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amproud1
}
\details{
Question How proud are you of being American? (Please, select one response)
}
\section{Values}{

\itemize{
\item \code{1} very proud
\item \code{2} somewhat proud
\item \code{3} not very proud
\item \code{4} not proud at all
\item \code{5} i am not american
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4895/vshow}{GSS Data Explorer page for \code{amproud1}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta2a}},
\code{\link{nafta3}}

Other United States: 
\code{\link{forcult}},
\code{\link{nafta2a}},
\code{\link{nafta3}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
