% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{amrank}
\alias{amrank}
\title{American quality of life vs other indus. nations}
\source{
General Social Survey https://gss.norc.org
}
\description{
amrank
}
\details{
Question Do you feel that the quality of life is better in America than in most other advanced industrial countries, about the same, or do you feel that people are better off in most other advanced industrial countries than they are in the United States?
}
\section{Values}{

\itemize{
\item \verb{[1]} better in america
\item \verb{[2]} about the same
\item \verb{[3]} better in oth adv cntry
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1597/vshow}{GSS Data Explorer page for \code{amrank}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab better in america \tab about the same \tab better in oth adv cntry \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 1155 \tab 220 \tab 61 \tab 35 \tab 3 \tab 2992 \cr
   Total \tab 1518 \tab 1155 \tab 220 \tab 61 \tab 35 \tab 3 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{meltpot}}
}
\concept{Multiculturalism}
\keyword{variable}
