% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amshamed}
\alias{amshamed}
\title{Agree there are things make me ashamed}
\source{
General Social Survey https://gss.norc.org
}
\description{
amshamed
}
\details{
Question How much do you agree or disagree with the following statements?
B. There are some things about America today that make me feel ashamed of America
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4835/vshow}{GSS Data Explorer page for \code{amshamed}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 222 \tab 647 \tab 210 \tab 181 \tab 68 \tab 25 \tab 14 \tab 2904 \cr
   2004 \tab 1596 \tab 219 \tab 455 \tab 228 \tab 197 \tab 103 \tab 13 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 233 \tab 591 \tab 202 \tab 154 \tab 58 \tab 33 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 674 \tab 1693 \tab 640 \tab 532 \tab 229 \tab 71 \tab 18 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
