% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{amtv}
\alias{amtv}
\title{Tv should give preference to american films}
\source{
General Social Survey https://gss.norc.org
}
\description{
amtv
}
\details{
Question How much do you agree or disagree with the following statements?
F. American television should give preference to American films and programs
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4855/vshow}{GSS Data Explorer page for \code{amtv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 124 \tab 359 \tab 421 \tab 317 \tab 77 \tab 45 \tab 24 \tab 2904 \cr
   2004 \tab 1596 \tab 105 \tab 292 \tab 350 \tab 362 \tab 79 \tab 27 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 70 \tab 286 \tab 370 \tab 379 \tab 110 \tab 55 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 299 \tab 937 \tab 1141 \tab 1058 \tab 266 \tab 127 \tab 29 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
