% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{aneseligible}
\alias{aneseligible}
\title{Anes follow-on eligible}
\source{
General Social Survey https://gss.norc.org
}
\description{
aneseligible
}
\details{
Question FLAG INDICATING WHETHER OR NOT THE RESPONDENT WAS ELIGIBLE FOR THE ANES
}
\section{Values}{

\itemize{
\item \verb{[0]} not eligible for follow-on
\item \verb{[1]} eligible for follow-on
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8272/vshow}{GSS Data Explorer page for \code{aneseligible}}.

Counts by year:\tabular{lllll}{
   year \tab not available in this year \tab not eligible for follow-on \tab eligible for follow-on \tab Total \cr
   2024 \tab - \tab 1777 \tab 1532 \tab 3309 \cr
   Total \tab 0 \tab 1777 \tab 1532 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{anespost_status}},
\code{\link{anespre_status}},
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovote24a}},
\code{\link{whovotets}}

Other ANES: 
\code{\link{anespost_status}},
\code{\link{anespre_status}},
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovote24a}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
