% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{anespost_status}
\alias{anespost_status}
\title{Anes post-election survey status}
\source{
General Social Survey https://gss.norc.org
}
\description{
anespost_status
}
\details{
Question FLAG INDICATING WHETHER OR NOT THE RESPONDENT PARTICIPATED IN THE ANES POST-ELECTION SURVEY
}
\section{Values}{

\itemize{
\item \verb{[1]} completed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8270/vshow}{GSS Data Explorer page for \code{anespost_status}}.

Counts by year:\tabular{lllll}{
   year \tab not available in this year \tab completed \tab iap \tab Total \cr
   2024 \tab - \tab 807 \tab 2502 \tab 3309 \cr
   Total \tab 0 \tab 807 \tab 2502 \tab 3309 \cr
}
}

\concept{ANES}
\keyword{variable}
