% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{anespre_status}
\alias{anespre_status}
\title{Anes pre-election survey status}
\source{
General Social Survey https://gss.norc.org
}
\description{
anespre_status
}
\details{
Question FLAG INDICATING WHETHER OR NOT THE RESPONDENT PARTICIPATED IN THE ANES PRE-ELECTION SURVEY
}
\section{Values}{

\itemize{
\item \verb{[1]} completed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8271/vshow}{GSS Data Explorer page for \code{anespre_status}}.

Counts by year:\tabular{lllll}{
   year \tab not available in this year \tab completed \tab iap \tab Total \cr
   2024 \tab - \tab 987 \tab 2322 \tab 3309 \cr
   Total \tab 0 \tab 987 \tab 2322 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{aneseligible}},
\code{\link{ratechall1ts}},
\code{\link{ratechall2ts}},
\code{\link{ratechall3ts}},
\code{\link{rateincumbts}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovotets}}

Other ANES: 
\code{\link{aneseligible}},
\code{\link{ratechall1ts}},
\code{\link{ratechall2ts}},
\code{\link{ratechall3ts}},
\code{\link{rateincumbts}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
