% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_185.R
\name{angryev}
\alias{angryev}
\title{Was an event recalled?}
\source{
General Social Survey https://gss.norc.org
}
\description{
angryev
}
\details{
Question Not enough information to code, not clear enough, not related to above dimensions
E. (CODE WITHOUT ASKING) WAS AN EVENT RECALLED?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1739/vshow}{GSS Data Explorer page for \code{angryev}}.
}

\seealso{
Other Emotions: 
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{beplesnt}},
\code{\link{madat1}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{whnangry}}
}
\concept{Emotions}
\keyword{variable}
