% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_185.R
\name{angryfam}
\alias{angryfam}
\title{Last time r felt really angry}
\source{
General Social Survey https://gss.norc.org
}
\description{
angryfam
}
\details{
Question Within the past month, think about the last time you felt really angry, irritated or annoyed. Could you describe in a couple of sentences what made you feel that way -- what the situation was? (INTERVIEWER: IF RESPONDENT CANNOT RECOVER AN EVENT, PROBE WITH "You haven't felt irritated or annoyed in the past month?")
B. Involved Family
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1736/vshow}{GSS Data Explorer page for \code{angryfam}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab Total \cr
   1996 \tab 1779 \tab 318 \tab 713 \tab 94 \tab 2904 \cr
   Total \tab 1779 \tab 318 \tab 713 \tab 94 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angryev}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{beplesnt}},
\code{\link{madat1}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{whnangry}}
}
\concept{Emotions}
\keyword{variable}
