% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_185.R
\name{angrygvt}
\alias{angrygvt}
\title{Last time r felt really angry}
\source{
General Social Survey https://gss.norc.org
}
\description{
angrygvt
}
\details{
Question Within the past month, think about the last time you felt really angry, irritated or annoyed. Could you describe in a couple of sentences what made you feel that way -- what the situation was? (INTERVIEWER: IF RESPONDENT CANNOT RECOVER AN EVENT, PROBE WITH "You haven't felt irritated or annoyed in the past month?")
C. Involved Government, Government officials
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1737/vshow}{GSS Data Explorer page for \code{angrygvt}}.
}

\seealso{
Other Emotions: 
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{beplesnt}},
\code{\link{madat1}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{whnangry}}
}
\concept{Emotions}
\keyword{variable}
