% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{anomia1}
\alias{anomia1}
\title{Next to health, money is most important}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia1
}
\details{
Question Now I'm going to read you several more statements. Some people agree with a statement, others disagree. As I read each one, tell me whether you more or less agree with it, or more or less disagree.
A. Next to health, money is the most important thing in life.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1976 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/537/vshow}{GSS Data Explorer page for \code{anomia1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 498 \tab 999 \tab 7 \tab - \tab 1504 \cr
   1974 \tab - \tab 424 \tab 1049 \tab 9 \tab 2 \tab 1484 \cr
   1976 \tab - \tab 501 \tab 992 \tab 6 \tab - \tab 1499 \cr
   Total \tab 0 \tab 1423 \tab 3040 \tab 22 \tab 2 \tab 4487 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{cigweek}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}
}
\concept{Core}
\keyword{variable}
