% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{anomia2}
\alias{anomia2}
\title{Wonder if anything is worthwhile}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia2
}
\details{
Question Now I'm going to read you several more statements. Some people agree with a statement, others disagree. As I read each one, tell me whether you more or less agree with it, or more or less disagree.
B. You sometimes can't help wondering whether anything is worthwhile any more.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1976 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/538/vshow}{GSS Data Explorer page for \code{anomia2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 604 \tab 885 \tab 12 \tab 3 \tab 1504 \cr
   1974 \tab - \tab 641 \tab 828 \tab 10 \tab 5 \tab 1484 \cr
   1976 \tab - \tab 596 \tab 881 \tab 17 \tab 5 \tab 1499 \cr
   Total \tab 0 \tab 1841 \tab 2594 \tab 39 \tab 13 \tab 4487 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{cigweek}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}
}
\concept{Core}
\keyword{variable}
