% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{anomia3}
\alias{anomia3}
\title{No right & wrong ways to make money}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia3
}
\details{
Question Now I'm going to read you several more statements. Some people agree with a statement, others disagree. As I read each one, tell me whether you more or less agree with it, or more or less disagree.
C. To make money, there are no right and wrong ways any more, only easy and hard ways.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1976 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/539/vshow}{GSS Data Explorer page for \code{anomia3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1973 \tab - \tab 376 \tab 1094 \tab 30 \tab 4 \tab - \tab 1504 \cr
   1974 \tab - \tab 311 \tab 1133 \tab 33 \tab 7 \tab - \tab 1484 \cr
   1976 \tab - \tab 395 \tab 1067 \tab 34 \tab 3 \tab - \tab 1499 \cr
   Total \tab 0 \tab 1082 \tab 3294 \tab 97 \tab 14 \tab 0 \tab 4487 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{cigweek}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}

Other Anomia: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}}

Other Split Ballots: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}}
}
\concept{Anomia}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
