% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{anomia4}
\alias{anomia4}
\title{Live only for today}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia4
}
\details{
Question Now I'm going to read you several more statements. Some people agree with a statement, others disagree. As I read each one, tell me whether you more or less agree with it, or more or less disagree.
D. Nowadays, a person has to live pretty much for today and let tomorrow take care of itself. Do you more or less agree with that, or more or less disagree?
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1976 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/540/vshow}{GSS Data Explorer page for \code{anomia4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 651 \tab 846 \tab 7 \tab - \tab 1504 \cr
   1974 \tab - \tab 646 \tab 822 \tab 14 \tab 2 \tab 1484 \cr
   1976 \tab - \tab 698 \tab 788 \tab 12 \tab 1 \tab 1499 \cr
   Total \tab 0 \tab 1995 \tab 2456 \tab 33 \tab 3 \tab 4487 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{cigweek}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}
}
\concept{Core}
\keyword{variable}
