% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{anomia5}
\alias{anomia5}
\title{Lot of the average man getting worse}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia5
}
\details{
Question Now I'm going to read you several more statements. Some people agree with a statement, others disagree. As I read each one, tell me whether you more or less agree with it, or more or less disagree.
E. In spite of what some people say, the lot (situation/condition) of the average man is getting worse, not better.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/541/vshow}{GSS Data Explorer page for \code{anomia5}}.
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia6}},
\code{\link{cigweek}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}
}
\concept{Core}
\keyword{variable}
