% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{anomia6}
\alias{anomia6}
\title{Not fair to bring child into world}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia6
}
\details{
Question Now I'm going to read you several more statements. Some people agree with a statement, others disagree. As I read each one, tell me whether you more or less agree with it, or more or less disagree.
F. It's hardly fair to bring a child into the world with the way things look for the future.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/542/vshow}{GSS Data Explorer page for \code{anomia6}}.
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{cigweek}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}

Other Anomia: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}}

Other Split Ballots: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}}
}
\concept{Anomia}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
