% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{anomia8}
\alias{anomia8}
\title{Don't know whom to trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia8
}
\details{
Question Now I''m going to read you several more statements. (See REMARKS.)
B. These days a person doesn''t really know whom he can count on.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1976 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/544/vshow}{GSS Data Explorer page for \code{anomia8}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 965 \tab 523 \tab 14 \tab 2 \tab 1504 \cr
   1974 \tab - \tab 1114 \tab 357 \tab 11 \tab 2 \tab 1484 \cr
   1976 \tab - \tab 1121 \tab 358 \tab 17 \tab 3 \tab 1499 \cr
   Total \tab 0 \tab 3200 \tab 1238 \tab 42 \tab 7 \tab 4487 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia7}},
\code{\link{anomia9}},
\code{\link{jobfind}},
\code{\link{jobinc}},
\code{\link{joblose}},
\code{\link{jobsec}},
\code{\link{richwork}},
\code{\link{satjob}}
}
\concept{Core}
\keyword{variable}
