% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{anomia9}
\alias{anomia9}
\title{Most don't care what happens to others}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia9
}
\details{
Question Now I''m going to read you several more statements. (See REMARKS.)
C. Most people don''t really care what happens to the next fellow.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1976 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/545/vshow}{GSS Data Explorer page for \code{anomia9}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 717 \tab 765 \tab 20 \tab 2 \tab 1504 \cr
   1974 \tab - \tab 867 \tab 596 \tab 18 \tab 3 \tab 1484 \cr
   1976 \tab - \tab 877 \tab 603 \tab 16 \tab 3 \tab 1499 \cr
   Total \tab 0 \tab 2461 \tab 1964 \tab 54 \tab 8 \tab 4487 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{jobfind}},
\code{\link{jobinc}},
\code{\link{joblose}},
\code{\link{jobsec}},
\code{\link{richwork}},
\code{\link{satjob}}
}
\concept{Core}
\keyword{variable}
