% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{anrights}
\alias{anrights}
\title{Animals have rights too}
\source{
General Social Survey https://gss.norc.org
}
\description{
anrights
}
\details{
Question And please check one box for each of these statements to show how much you agree or disagree with it.
C. Animals should have the same moral rights that human beings do.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4727/vshow}{GSS Data Explorer page for \code{anrights}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 95 \tab 373 \tab 285 \tab 516 \tab 198 \tab 36 \tab 54 \tab 1606 \cr
   1994 \tab 1606 \tab 73 \tab 286 \tab 257 \tab 447 \tab 241 \tab 38 \tab 44 \tab 2992 \cr
   Total \tab 1655 \tab 168 \tab 659 \tab 542 \tab 963 \tab 439 \tab 74 \tab 98 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grnecon}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scifaith}},
\code{\link{scigrn}},
\code{\link{sciworse}}
}
\concept{ISSP Environment}
\keyword{variable}
