% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{arrest}
\alias{arrest}
\title{Ever picked up or charged by police}
\source{
General Social Survey https://gss.norc.org
}
\description{
arrest
}
\details{
Question A. Were you ever picked up, or charged, by the police, for any (other) reason whether or not you were guilty?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/685/vshow}{GSS Data Explorer page for \code{arrest}}.
}

\seealso{
Other Core: 
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}},
\code{\link{tvhours}}

Other Arrest (R's): 
\code{\link{ticket}}

Other Police: 
\code{\link{lockedup}},
\code{\link{ticket}}
}
\concept{Arrest (R's)}
\concept{Core}
\concept{Police}
\keyword{variable}
