% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_235.R
\name{artatt}
\alias{artatt}
\title{Attended exhibit alone or with others}
\source{
General Social Survey https://gss.norc.org
}
\description{
artatt
}
\details{
Question You said that in the last 12 months you attended an exhibit. I'd like to ask you a few questions about the last time you did this type of activity.
Did you attend this exhibit alone, with one other person, or with two or more other people?
}
\section{Values}{

\itemize{
\item \verb{[1]} alone
\item \verb{[2]} with one other person
\item \verb{[3]} with two or more other people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6231/vshow}{GSS Data Explorer page for \code{artatt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab alone \tab with one other person \tab with two or more other people \tab no answer \tab Total \cr
   2016 \tab 2546 \tab 29 \tab 150 \tab 134 \tab 8 \tab 2867 \cr
   Total \tab 2546 \tab 29 \tab 150 \tab 134 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt4}},
\code{\link{artatt5}},
\code{\link{artfree}},
\code{\link{artwhy1}},
\code{\link{artwhy2}},
\code{\link{prfmwhy0}}
}
\concept{Arts and Culture}
\keyword{variable}
