% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_235.R
\name{artatt4}
\alias{artatt4}
\title{Attended with relative}
\source{
General Social Survey https://gss.norc.org
}
\description{
artatt4
}
\details{
Question Relative
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6235/vshow}{GSS Data Explorer page for \code{artatt4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 2575 \tab 74 \tab 210 \tab 8 \tab 2867 \cr
   Total \tab 2575 \tab 74 \tab 210 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt}},
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt5}},
\code{\link{artfree}},
\code{\link{artwhy1}},
\code{\link{artwhy2}},
\code{\link{prfmwhy0}}
}
\concept{Arts and Culture}
\keyword{variable}
