% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_233.R
\name{artexbt}
\alias{artexbt}
\title{Did r go to an art exhibit in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
artexbt
}
\details{
Question During the last 12 months, did you go to an art exhibit, such as paintings, sculpture, textiles, graphic design, or photography?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6210/vshow}{GSS Data Explorer page for \code{artexbt}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{prfmatt}},
\code{\link{prfmatt1}},
\code{\link{prfmatt2}},
\code{\link{prfmatt3}},
\code{\link{prfmatt4}},
\code{\link{prfmatt5}},
\code{\link{prfmdan}},
\code{\link{prfmmus}},
\code{\link{prfmthe}}
}
\concept{Arts and Culture}
\keyword{variable}
