% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{artgod}
\alias{artgod}
\title{Thru art and music we come closer to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
artgod
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example . . .
A. Through such things as art and music, we come closer to God.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2124/vshow}{GSS Data Explorer page for \code{artgod}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 161 \tab 741 \tab 311 \tab 80 \tab 134 \tab 18 \tab 2832 \cr
   Total \tab 1387 \tab 161 \tab 741 \tab 311 \tab 80 \tab 134 \tab 18 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{artists}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{natarts}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion and Culture}
\keyword{variable}
