% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{artmostoth}
\alias{artmostoth}
\title{Most imp factor in not attending art exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
artmostoth
}
\details{
Question Of these factors, which one was the most important in your decision not to attend?
}
\section{Values}{

\itemize{
\item \code{1} cost too much
\item \code{2} programs or events were not of interest
\item \code{3} too difficult to get there
\item \code{4} could not find anyone to go with
\item \code{5} could not find the time
\item \code{6} did not want to go to that location
\item \code{7} work conflict
\item \code{8} ill, too tired
\item \code{10} no childcare
\item \code{11} sold out
\item \code{12} forgot, didn't act in time
\item \code{13} away, other conflict
\item \code{14} taking care of family member
\item \code{20} some other reason not shown
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6265/vshow}{GSS Data Explorer page for \code{artmostoth}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt}},
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt4}},
\code{\link{artatt5}},
\code{\link{artcost}},
\code{\link{artexbt}},
\code{\link{artfree}},
\code{\link{artgst}},
\code{\link{artint}},
\code{\link{artloc}},
\code{\link{artothr}},
\code{\link{arttime}},
\code{\link{arttrvl}},
\code{\link{artwhy1}},
\code{\link{artwhy2}},
\code{\link{artwhy3}},
\code{\link{artwhy4}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo}},
\code{\link{nogo1}},
\code{\link{prfmatt}},
\code{\link{prfmatt1}},
\code{\link{prfmatt2}},
\code{\link{prfmatt3}},
\code{\link{prfmatt4}},
\code{\link{prfmatt5}},
\code{\link{prfmcost}},
\code{\link{prfmdan}},
\code{\link{prfmfree}},
\code{\link{prfmgst}},
\code{\link{prfmint}},
\code{\link{prfmloc}},
\code{\link{prfmmostoth}},
\code{\link{prfmmus}},
\code{\link{prfmnce}},
\code{\link{prfmothr}},
\code{\link{prfmthe}},
\code{\link{prfmtime}},
\code{\link{prfmtrvl}},
\code{\link{prfmwhy0}},
\code{\link{prfmwhy1}},
\code{\link{prfmwhy2}},
\code{\link{prfmwhy3}},
\code{\link{prfmwhy4}},
\code{\link{prfmwhy5}},
\code{\link{prfmwhy6}},
\code{\link{prfmwhy7}},
\code{\link{prfmwhy8}},
\code{\link{prfmwhy9}}
}
\concept{Arts and Culture}
\keyword{variable}
