% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{artmus1}
\alias{artmus1}
\title{Use web to learn about upcoming exhibits}
\source{
General Social Survey https://gss.norc.org
}
\description{
artmus1
}
\details{
Question B. Visit the Web site of an art museum or art gallery in order to learn about upcoming exhibits or to get information before visiting that museum?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2474/vshow}{GSS Data Explorer page for \code{artmus1}}.
}

\seealso{
Other Information Society: 
\code{\link{artinfo}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{musicnew}},
\code{\link{wwwart}},
\code{\link{wwwworld}}

Other Computers And Internet: 
\code{\link{artinfo}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{musicnew}},
\code{\link{wwwart}},
\code{\link{wwwworld}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
