% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{artothr}
\alias{artothr}
\title{Other reason why did not go}
\source{
General Social Survey https://gss.norc.org
}
\description{
artothr
}
\details{
Question Some other reason not shown.
}
\section{Values}{

\itemize{
\item \verb{[1]} a reason
\item \verb{[2]} not a reason
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6262/vshow}{GSS Data Explorer page for \code{artothr}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab a reason \tab not a reason \tab don't know \tab Total \cr
   2016 \tab 2710 \tab 45 \tab 111 \tab 1 \tab 2867 \cr
   Total \tab 2710 \tab 45 \tab 111 \tab 1 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artgst}},
\code{\link{artloc}},
\code{\link{artmostoth}},
\code{\link{arttime}},
\code{\link{prfmmostoth}}
}
\concept{Arts and Culture}
\keyword{variable}
