% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{artschat}
\alias{artschat}
\title{Use chatrooms about arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
artschat
}
\details{
Question Considering all these arts topics, do you participate in chat rooms, Usenet groups, discussion forums, bulletIn boards, or similar interactive services devoted to these topics?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2453/vshow}{GSS Data Explorer page for \code{artschat}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2075 \tab 72 \tab 587 \tab 83 \tab 2817 \cr
   2002 \tab 2043 \tab 67 \tab 609 \tab 46 \tab 2765 \cr
   Total \tab 4118 \tab 139 \tab 1196 \tab 129 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
