% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{artschng}
\alias{artschng}
\title{Web change mind about arts on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
artschng
}
\details{
Question Now I would like you to think about the last time you used the Web to get information or exchange views about music, film, the visual arts, dance, theater, museums, literature, or related topics.
B. Did your Web experience change your tastes or opinions in any way?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2455/vshow}{GSS Data Explorer page for \code{artschng}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2355 \tab 61 \tab 278 \tab 123 \tab - \tab 2817 \cr
   2002 \tab 2284 \tab 73 \tab 360 \tab 46 \tab 2 \tab 2765 \cr
   Total \tab 4639 \tab 134 \tab 638 \tab 169 \tab 2 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
