% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{artseals}
\alias{artseals}
\title{Effect of global warming: arctic seals are threatened}
\source{
General Social Survey https://gss.norc.org
}
\description{
artseals
}
\details{
Question Please look at Card B24. Scientists predict that global warming may soon have big effects on the polar regions. I will describe some of these possible effects and, for each one, please say whether it would bother you a great deal, some, a little, or not at all if it actually happened.
C. Arctic seals may be threatened.
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} some
\item \code{3} a little
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3440/vshow}{GSS Data Explorer page for \code{artseals}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab some \tab a little \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 808 \tab 658 \tab 222 \tab 113 \tab 60 \tab 3 \tab 4510 \cr
   Total \tab 2646 \tab 808 \tab 658 \tab 222 \tab 113 \tab 60 \tab 3 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
