% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{artshr}
\alias{artshr}
\title{How many hours at arts websites}
\source{
General Social Survey https://gss.norc.org
}
\description{
artshr
}
\details{
Question In a typical month, about how many minutes or hours do you use the Web to explore sites that have to do with music, film, the visual arts, dance, theater, museums, literature or related topics?
B. Hours per month?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2452/vshow}{GSS Data Explorer page for \code{artshr}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 120 \tab 14 \tab 15 \tab 2 \tab 20 \tab 24 \tab 25 \tab 3 \tab 30 \tab 392 \tab 4 \tab 40 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 8 \tab 80 \tab 9 \tab don't know \tab no answer \tab 17 \tab 32 \tab Total \cr
   2000 \tab 2075 \tab 373 \tab 89 \tab 25 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 58 \tab 6 \tab 1 \tab 2 \tab 24 \tab 2 \tab 1 \tab 17 \tab 3 \tab 24 \tab 2 \tab 13 \tab 2 \tab 6 \tab 4 \tab 1 \tab 1 \tab 3 \tab 76 \tab - \tab - \tab 2817 \cr
   2002 \tab 2043 \tab 359 \tab 98 \tab 14 \tab 1 \tab 2 \tab - \tab 1 \tab 11 \tab 59 \tab 13 \tab - \tab 2 \tab 38 \tab 5 \tab - \tab 13 \tab 1 \tab 35 \tab 1 \tab 10 \tab - \tab 3 \tab 6 \tab - \tab 1 \tab 1 \tab 46 \tab 1 \tab 1 \tab 2765 \cr
   Total \tab 4118 \tab 732 \tab 187 \tab 39 \tab 2 \tab 5 \tab 1 \tab 2 \tab 14 \tab 117 \tab 19 \tab 1 \tab 4 \tab 62 \tab 7 \tab 1 \tab 30 \tab 4 \tab 59 \tab 3 \tab 23 \tab 2 \tab 9 \tab 10 \tab 1 \tab 2 \tab 4 \tab 122 \tab 1 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
