% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{artsmin}
\alias{artsmin}
\title{How many minutes at arts websites}
\source{
General Social Survey https://gss.norc.org
}
\description{
artsmin
}
\details{
Question In a typical month, about how many minutes or hours do you use the Web to explore sites that have to do with music, film, the visual arts, dance, theater, museums, literature or related topics?
A. Minutes per month?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2451/vshow}{GSS Data Explorer page for \code{artsmin}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 15 \tab 16 \tab 20 \tab 30 \tab 32 \tab 40 \tab 45 \tab 5 \tab 50 \tab 59 \tab don't know \tab no answer \tab 19 \tab 3 \tab Total \cr
   2000 \tab 2075 \tab 551 \tab 13 \tab 17 \tab 1 \tab 18 \tab 57 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 74 \tab - \tab - \tab 2817 \cr
   2002 \tab 2043 \tab 562 \tab 16 \tab 17 \tab - \tab 14 \tab 56 \tab - \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 46 \tab 1 \tab 1 \tab 2765 \cr
   Total \tab 4118 \tab 1113 \tab 29 \tab 34 \tab 1 \tab 32 \tab 113 \tab 1 \tab 2 \tab 3 \tab 8 \tab 1 \tab 1 \tab 4 \tab 120 \tab 1 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
