% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{artsnew}
\alias{artsnew}
\title{R learn something new on web about arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
artsnew
}
\details{
Question Now I would like you to think about the last time you used the Web to get information or exchange views about music, film, the visual arts, dance, theater, museums, literature, or related topics.
A. On that occasion, did you come across or learn about anything new with regard to any of these topics?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{8} cant recall
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2454/vshow}{GSS Data Explorer page for \code{artsnew}}.
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
