% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_654.R
\name{artsout}
\alias{artsout}
\title{Outdoor art activity in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
artsout
}
\details{
Question Thinking about all the types of events you said you attended, were any of those activities outdoors?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7869/vshow}{GSS Data Explorer page for \code{artsout}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab Total \cr
   2022 \tab 2929 \tab 368 \tab 244 \tab 3 \tab 3544 \cr
   Total \tab 2929 \tab 368 \tab 244 \tab 3 \tab 3544 \cr
}
}

\seealso{
Other NEA: 
\code{\link{yrarmus}},
\code{\link{yrarpo}},
\code{\link{yrclass}},
\code{\link{yrcreat}},
\code{\link{yrmovie}},
\code{\link{yrrdg}},
\code{\link{yrstmus}},
\code{\link{yrstpo}},
\code{\link{yrtour}}
}
\concept{NEA}
\keyword{variable}
