% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_237.R
\name{arttrvl}
\alias{arttrvl}
\title{Did not go because too dificult to get there}
\source{
General Social Survey https://gss.norc.org
}
\description{
arttrvl
}
\details{
Question Too difficult to get there.
}
\section{Values}{

\itemize{
\item \code{1} a reason
\item \code{2} not a reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6258/vshow}{GSS Data Explorer page for \code{arttrvl}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab a reason \tab not a reason \tab Total \cr
   2016 \tab 2710 \tab 47 \tab 110 \tab 2867 \cr
   Total \tab 2710 \tab 47 \tab 110 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artcost}},
\code{\link{artint}},
\code{\link{prfmgst}},
\code{\link{prfmint}},
\code{\link{prfmloc}},
\code{\link{prfmothr}},
\code{\link{prfmtime}},
\code{\link{prfmtrvl}}
}
\concept{Arts and Culture}
\keyword{variable}
