% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_235.R
\name{artwhy1}
\alias{artwhy1}
\title{Imp of low cost in decision to attend exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
artwhy1
}
\details{
Question People attend exhibits for different reasons. How big a reason was low cost in your decision to attend this exhibit? Was it a major reason, a minor reason, or not a reason at all?     
SAY, "THIS CARD SHOWS SOME REASONS THAT PEOPLE ATTEND EXHIBITS."
}
\section{Values}{

\itemize{
\item \verb{[1]} major reason
\item \verb{[2]} minor reason
\item \verb{[3]} not a reason
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6238/vshow}{GSS Data Explorer page for \code{artwhy1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab major reason \tab minor reason \tab not a reason \tab no answer \tab Total \cr
   2016 \tab 2546 \tab 35 \tab 67 \tab 211 \tab 8 \tab 2867 \cr
   Total \tab 2546 \tab 35 \tab 67 \tab 211 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt}},
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt4}},
\code{\link{artatt5}},
\code{\link{artfree}},
\code{\link{artwhy2}},
\code{\link{prfmwhy0}}
}
\concept{Arts and Culture}
\keyword{variable}
