% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_236.R
\name{artwhy4}
\alias{artwhy4}
\title{Imp of wanting to celebrate cultural heritage in decision to attend exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
artwhy4
}
\details{
Question How big a reason was wanting to learn about or celebrate your or your family's cultural heritage in your decision to attend this exhibit? Was it a major reason, a minor reason, or not a reason at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} major reason
\item \verb{[2]} minor reason
\item \verb{[3]} not a reason
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6241/vshow}{GSS Data Explorer page for \code{artwhy4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab major reason \tab minor reason \tab not a reason \tab no answer \tab Total \cr
   2016 \tab 2546 \tab 29 \tab 49 \tab 235 \tab 8 \tab 2867 \cr
   Total \tab 2546 \tab 29 \tab 49 \tab 235 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artwhy3}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo}},
\code{\link{nogo1}},
\code{\link{prfmcost}}
}
\concept{Arts and Culture}
\keyword{variable}
