% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_236.R
\name{artwhy9}
\alias{artwhy9}
\title{Other major reason in decision to attend}
\source{
General Social Survey https://gss.norc.org
}
\description{
artwhy9
}
\details{
Question Please specify (Immediately after ARTWHY8)
}
\section{Values}{

\itemize{
\item \verb{[1]} gift, free
\item \verb{[2]} school related
\item \verb{[3]} wanted to see specific show/exhibit
\item \verb{[4]} variety/something different/new experience
\item \verb{[5]} win meal, drinks
\item \verb{[6]} entertainment, enjoyment, have fun
\item \verb{[7]} religious
\item \verb{[8]} for children, educational
\item \verb{[10]} support family/friend
\item \verb{[11]} celebrate birthday, anniversary
\item \verb{[12]} night out, date
\item \verb{[13]} meet, support artist
\item \verb{[14]} work related
\item \verb{[20]} miscellaneous other reasons
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6246/vshow}{GSS Data Explorer page for \code{artwhy9}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab school related \tab wanted to see specific show/exhibit \tab entertainment, enjoyment, have fun \tab support family/friend \tab celebrate birthday, anniversary \tab night out, date \tab meet, support artist \tab work related \tab miscellaneous other reasons \tab no answer \tab Total \cr
   2016 \tab 2798 \tab 5 \tab 1 \tab 1 \tab 17 \tab 2 \tab 2 \tab 2 \tab 2 \tab 29 \tab 8 \tab 2867 \cr
   Total \tab 2798 \tab 5 \tab 1 \tab 1 \tab 17 \tab 2 \tab 2 \tab 2 \tab 2 \tab 29 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artwhy3}},
\code{\link{artwhy4}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{nogo}},
\code{\link{nogo1}},
\code{\link{prfmcost}}
}
\concept{Arts and Culture}
\keyword{variable}
