% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{askfinan}
\alias{askfinan}
\title{Govt investigate workers finances for security}
\source{
General Social Survey https://gss.norc.org
}
\description{
askfinan
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
A. Financial and credit history
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5208/vshow}{GSS Data Explorer page for \code{askfinan}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 801 \tab 388 \tab 159 \tab 89 \tab 26 \tab 11 \tab 2992 \cr
   1996 \tab 1444 \tab 743 \tab 413 \tab 152 \tab 113 \tab 35 \tab 4 \tab 2904 \cr
   1998 \tab 1445 \tab 612 \tab 416 \tab 185 \tab 121 \tab 45 \tab 8 \tab 2832 \cr
   2000 \tab 1398 \tab 693 \tab 393 \tab 159 \tab 134 \tab 33 \tab 7 \tab 2817 \cr
   Total \tab 5805 \tab 2849 \tab 1610 \tab 655 \tab 457 \tab 139 \tab 30 \tab 11545 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{secdocs}},
\code{\link{sectech}}
}
\concept{National Security}
\keyword{variable}
