% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{askforgn}
\alias{askforgn}
\title{Govt investigate workers foreign family}
\source{
General Social Survey https://gss.norc.org
}
\description{
askforgn
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
E. Foreign relatives and friends
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5212/vshow}{GSS Data Explorer page for \code{askforgn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 682 \tab 451 \tab 182 \tab 110 \tab 38 \tab 11 \tab 2992 \cr
   1996 \tab 1444 \tab 703 \tab 441 \tab 188 \tab 76 \tab 47 \tab 5 \tab 2904 \cr
   1998 \tab 1445 \tab 628 \tab 438 \tab 192 \tab 77 \tab 42 \tab 10 \tab 2832 \cr
   2000 \tab 1398 \tab 701 \tab 383 \tab 176 \tab 108 \tab 42 \tab 9 \tab 2817 \cr
   Total \tab 5805 \tab 2714 \tab 1713 \tab 738 \tab 371 \tab 169 \tab 35 \tab 11545 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{secdocs}},
\code{\link{sectech}}
}
\concept{National Security}
\keyword{variable}
