% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{askmentl}
\alias{askmentl}
\title{Govt investigate workers mental health}
\source{
General Social Survey https://gss.norc.org
}
\description{
askmentl
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
D. Mental health history
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5211/vshow}{GSS Data Explorer page for \code{askmentl}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 1109 \tab 273 \tab 48 \tab 12 \tab 21 \tab 11 \tab 2992 \cr
   1996 \tab 1444 \tab 1143 \tab 240 \tab 38 \tab 7 \tab 27 \tab 5 \tab 2904 \cr
   1998 \tab 1445 \tab 977 \tab 323 \tab 36 \tab 14 \tab 29 \tab 8 \tab 2832 \cr
   2000 \tab 1398 \tab 1035 \tab 272 \tab 51 \tab 22 \tab 31 \tab 8 \tab 2817 \cr
   Total \tab 5805 \tab 4264 \tab 1108 \tab 173 \tab 55 \tab 108 \tab 32 \tab 11545 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{rptcowrk}},
\code{\link{secdocs}},
\code{\link{sectech}}
}
\concept{National Security}
\keyword{variable}
