% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{asksexor}
\alias{asksexor}
\title{Govt investigate workers sexual orientation}
\source{
General Social Survey https://gss.norc.org
}
\description{
asksexor
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
G. Sexual orientation
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5214/vshow}{GSS Data Explorer page for \code{asksexor}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 421 \tab 274 \tab 362 \tab 364 \tab 41 \tab 12 \tab 2992 \cr
   1996 \tab 1444 \tab 416 \tab 273 \tab 390 \tab 327 \tab 50 \tab 4 \tab 2904 \cr
   1998 \tab 1445 \tab 358 \tab 258 \tab 349 \tab 367 \tab 46 \tab 9 \tab 2832 \cr
   2000 \tab 1398 \tab 394 \tab 228 \tab 362 \tab 384 \tab 42 \tab 9 \tab 2817 \cr
   Total \tab 5805 \tab 1589 \tab 1033 \tab 1463 \tab 1442 \tab 179 \tab 34 \tab 11545 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{secterr}}
}
\concept{National Security}
\keyword{variable}
