% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_257.R
\name{asnchng}
\alias{asnchng}
\title{Percentage change population of asians in 25 yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
asnchng
}
\details{
Question In the next 25 years, which of these groups do you think will increase their share of the population by a lot, which will increase by a little, which will stay the same, which will decrease by a little, and which will decrease by a lot?
E. Asian Americans?
}
\section{Values}{

\itemize{
\item \code{1} increase a lot
\item \code{2} increase by a little
\item \code{3} stay the same
\item \code{4} decrease by a little
\item \code{5} decrease by a lot
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2321/vshow}{GSS Data Explorer page for \code{asnchng}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increase a lot \tab increase by a little \tab stay the same \tab decrease by a little \tab decrease by a lot \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 269 \tab 607 \tab 341 \tab 40 \tab 17 \tab 110 \tab 14 \tab 2817 \cr
   Total \tab 1419 \tab 269 \tab 607 \tab 341 \tab 40 \tab 17 \tab 110 \tab 14 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{ethchng}},
\code{\link{hispwork}},
\code{\link{knwwht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
