% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_260.R
\name{asncls}
\alias{asncls}
\title{Does r feel close to asians}
\source{
General Social Survey https://gss.norc.org
}
\description{
asncls
}
\details{
Question Now, I’m going to ask some questions about these people you personally know.
E. Are any of these Asians people that you feel close to?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2359/vshow}{GSS Data Explorer page for \code{asncls}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispwrk}},
\code{\link{mostcom}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
