% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{athletic}
\alias{athletic}
\title{R is an athletic person}
\source{
General Social Survey https://gss.norc.org
}
\description{
athletic
}
\details{
Question For each of the following, please indicate how well the description applies to you by circling one number:
C. An athletic person. Is this …...
}
\section{Values}{

\itemize{
\item \verb{[1]} a very good description
\item \verb{[2]} a good description
\item \verb{[3]} a fair description
\item \verb{[4]} not a very good description
\item \verb{[5]} not a good description at all
\item \verb{[9]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3049/vshow}{GSS Data Explorer page for \code{athletic}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a very good description \tab a good description \tab a fair description \tab not a very good description \tab not a good description at all \tab don't know \tab no answer \tab Total \cr
   2004 \tab 432 \tab 255 \tab 429 \tab 761 \tab 633 \tab 295 \tab 1 \tab 6 \tab 2812 \cr
   Total \tab 432 \tab 255 \tab 429 \tab 761 \tab 633 \tab 295 \tab 1 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{owndoing}},
\code{\link{sadblue}},
\code{\link{selfish}}
}
\concept{Genes and the Environment}
\keyword{variable}
