% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{attend}
\alias{attend}
\title{How often r attends religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
attend
}
\details{
Question How often do you attend religious services? (USE CATEGORIES AS PROBES, IF NECESSARY.)
}
\section{Values}{

\itemize{
\item \verb{[0]} never
\item \verb{[1]} less than once a year
\item \verb{[2]} about once or twice a year
\item \verb{[3]} several times a year
\item \verb{[4]} about once a month
\item \verb{[5]} 2-3 times a month
\item \verb{[6]} nearly every week
\item \verb{[7]} every week
\item \verb{[8]} several times a week
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/294/vshow}{GSS Data Explorer page for \code{attend}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab never \tab less than once a year \tab about once or twice a year \tab several times a year \tab about once a month \tab 2-3 times a month \tab nearly every week \tab every week \tab several times a week \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 150 \tab 137 \tab 178 \tab 226 \tab 108 \tab 142 \tab 98 \tab 460 \tab 101 \tab 13 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 204 \tab 117 \tab 194 \tab 226 \tab 93 \tab 125 \tab 116 \tab 308 \tab 112 \tab 9 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 181 \tab 103 \tab 226 \tab 188 \tab 118 \tab 129 \tab 86 \tab 333 \tab 117 \tab 3 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 216 \tab 108 \tab 184 \tab 209 \tab 102 \tab 133 \tab 99 \tab 334 \tab 102 \tab 3 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 194 \tab 136 \tab 203 \tab 232 \tab 106 \tab 100 \tab 90 \tab 298 \tab 133 \tab 7 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 212 \tab 127 \tab 204 \tab 189 \tab 113 \tab 138 \tab 87 \tab 328 \tab 123 \tab 9 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 239 \tab 134 \tab 202 \tab 175 \tab 103 \tab 145 \tab 107 \tab 299 \tab 123 \tab 5 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 167 \tab 109 \tab 230 \tab 223 \tab 98 \tab 118 \tab 91 \tab 316 \tab 109 \tab 7 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 241 \tab 131 \tab 264 \tab 258 \tab 132 \tab 180 \tab 120 \tab 363 \tab 158 \tab 13 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 222 \tab 130 \tab 210 \tab 186 \tab 105 \tab 155 \tab 85 \tab 366 \tab 136 \tab 4 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 190 \tab 107 \tab 173 \tab 206 \tab 110 \tab 122 \tab 71 \tab 345 \tab 136 \tab 13 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 222 \tab 107 \tab 230 \tab 185 \tab 113 \tab 114 \tab 57 \tab 383 \tab 119 \tab 4 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 206 \tab 105 \tab 183 \tab 175 \tab 117 \tab 148 \tab 66 \tab 334 \tab 132 \tab 4 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 196 \tab 113 \tab 240 \tab 272 \tab 154 \tab 215 \tab 97 \tab 373 \tab 146 \tab 13 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 254 \tab 108 \tab 168 \tab 194 \tab 116 \tab 143 \tab 109 \tab 277 \tab 109 \tab 3 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 252 \tab 111 \tab 199 \tab 189 \tab 123 \tab 125 \tab 80 \tab 347 \tab 107 \tab 4 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 180 \tab 114 \tab 163 \tab 175 \tab 105 \tab 127 \tab 70 \tab 306 \tab 93 \tab 39 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 190 \tab 136 \tab 202 \tab 179 \tab 118 \tab 148 \tab 83 \tab 344 \tab 92 \tab 25 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 259 \tab 136 \tab 193 \tab 176 \tab 112 \tab 137 \tab 99 \tab 321 \tab 135 \tab 38 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 481 \tab 226 \tab 416 \tab 382 \tab 221 \tab 274 \tab 146 \tab 560 \tab 235 \tab 51 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 438 \tab 253 \tab 400 \tab 416 \tab 190 \tab 271 \tab 160 \tab 486 \tab 209 \tab 81 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 546 \tab 294 \tab 296 \tab 307 \tab 206 \tab 244 \tab 188 \tab 489 \tab 218 \tab 44 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 583 \tab 219 \tab 334 \tab 369 \tab 197 \tab 221 \tab 131 \tab 488 \tab 195 \tab 80 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 513 \tab 194 \tab 385 \tab 357 \tab 188 \tab 258 \tab 180 \tab 455 \tab 213 \tab 22 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 471 \tab 198 \tab 396 \tab 371 \tab 191 \tab 255 \tab 169 \tab 508 \tab 242 \tab 11 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1020 \tab 302 \tab 571 \tab 502 \tab 308 \tab 380 \tab 240 \tab 839 \tab 329 \tab 19 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 427 \tab 142 \tab 281 \tab 233 \tab 140 \tab 174 \tab 88 \tab 361 \tab 168 \tab 9 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 470 \tab 143 \tab 287 \tab 211 \tab 147 \tab 171 \tab 87 \tab 382 \tab 138 \tab 8 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 496 \tab 111 \tab 264 \tab 191 \tab 133 \tab 168 \tab 83 \tab 380 \tab 140 \tab 8 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 669 \tab 183 \tab 337 \tab 250 \tab 149 \tab 217 \tab 114 \tab 431 \tab 175 \tab 13 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 723 \tab 173 \tab 385 \tab 312 \tab 190 \tab 250 \tab 126 \tab 481 \tab 210 \tab 17 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 716 \tab 137 \tab 300 \tab 239 \tab 146 \tab 186 \tab 88 \tab 391 \tab 129 \tab 16 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1178 \tab 565 \tab 453 \tab 403 \tab 122 \tab 200 \tab 331 \tab 532 \tab 178 \tab - \tab 40 \tab 3 \tab 27 \tab 4032 \cr
   2022 \tab 1149 \tab 416 \tab 464 \tab 346 \tab 157 \tab 201 \tab 187 \tab 441 \tab 148 \tab - \tab 2 \tab 24 \tab 9 \tab 3544 \cr
   Total \tab 13855 \tab 5825 \tab 9415 \tab 8752 \tab 4831 \tab 6114 \tab 4029 \tab 13659 \tab 5210 \tab 595 \tab 42 \tab 27 \tab 36 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{fund}},
\code{\link{jew}},
\code{\link{jewaj}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{spattend}}
}
\concept{Core}
\keyword{variable}
