% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{attndgrp}
\alias{attndgrp}
\title{Has r attended group conferences}
\source{
General Social Survey https://gss.norc.org
}
\description{
attndgrp
}
\details{
Question Here is a list of things that members of organizations may participate in. Could you tell me whether or not you have ever done any of these things as a member of (GROUP IN MEMFRAT-MEMNUM OR 332)? (READ EACH ITEM. CIRCLE A NUMBER FOR EACH ACTIVITY).
D. Attended conferences or workshops
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1018/vshow}{GSS Data Explorer page for \code{attndgrp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1987 \tab 606 \tab 713 \tab 463 \tab 1 \tab 36 \tab 1819 \cr
   Total \tab 606 \tab 713 \tab 463 \tab 1 \tab 36 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{mostact}},
\code{\link{servegrp}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
