% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{attractd}
\alias{attractd}
\title{At what age were you first sexually attracted to same sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
attractd
}
\details{
Question At about what age were you first sexually attracted to someone of the same sex?
}
\section{Values}{

\itemize{
\item \code{1} under age 14
\item \code{2} age 14-18
\item \code{3} age 18-25
\item \code{4} age 26-35
\item \code{5} age 36-45
\item \code{6} age 46-55
\item \code{7} age 56-65
\item \code{8} age 66 or older
\item \code{9} i have never told anyone i have never been sexually attracted to someone of the same sex
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3908/vshow}{GSS Data Explorer page for \code{attractd}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab under age 14 \tab age 14-18 \tab age 18-25 \tab age 26-35 \tab age 36-45 \tab age 46-55 \tab i have never told anyone i have never been sexually attracted to someone of the same sex \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1879 \tab 30 \tab 23 \tab 13 \tab 3 \tab 1 \tab 4 \tab 20 \tab 4 \tab 46 \tab 2023 \cr
   Total \tab 1879 \tab 30 \tab 23 \tab 13 \tab 3 \tab 1 \tab 4 \tab 20 \tab 4 \tab 46 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{toldsxor}}
}
\concept{Sexual Orientation}
\keyword{variable}
