% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_539.R
\name{attrally}
\alias{attrally}
\title{Attended a political meeting or rally}
\source{
General Social Survey https://gss.norc.org
}
\description{
attrally
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the past year, whether you have done it in the more distant past, whether you have not done it but might do it, or have not done it and would never, under any circumstances, do it.
D. Attended a political meeting or rally
}
\section{Values}{

\itemize{
\item \verb{[1]} have done it in the past yr
\item \verb{[2]} have done it in the more distant past
\item \verb{[3]} have not done it but might do it
\item \verb{[4]} have not done it and would never do it
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4923/vshow}{GSS Data Explorer page for \code{attrally}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab have done it in the past yr \tab have done it in the more distant past \tab have not done it but might do it \tab have not done it and would never do it \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 183 \tab 293 \tab 585 \tab 406 \tab 1 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 100 \tab 269 \tab 479 \tab 395 \tab 19 \tab 2 \tab 2538 \cr
   Total \tab 2614 \tab 283 \tab 562 \tab 1064 \tab 801 \tab 20 \tab 6 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{avoidbuy}},
\code{\link{cntctgov}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{polfunds}},
\code{\link{racmeet}},
\code{\link{revmeet}},
\code{\link{signdpet}},
\code{\link{usemedia}}
}
\concept{ISSP Citizenship}
\keyword{variable}
