% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_149.R
\name{attsprts}
\alias{attsprts}
\title{Attended sports event in lst yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
attsprts
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
A. Attend an amateur or professional sports event.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1405/vshow}{GSS Data Explorer page for \code{attsprts}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 855 \tab 739 \tab 2 \tab 10 \tab 1606 \cr
   Total \tab 0 \tab 855 \tab 739 \tab 2 \tab 10 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{autorace}},
\code{\link{conrock}},
\code{\link{hvymetal}},
\code{\link{makeart}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{rap}},
\code{\link{reggae}},
\code{\link{visitart}}
}
\concept{Culture}
\keyword{variable}
