% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_582.R
\name{away2}
\alias{away2}
\title{Is 2nd person staying somewhere else now?}
\source{
General Social Survey https://gss.norc.org
}
\description{
away2
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
5. Is the second person staying somewhere else right now?
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5342/vshow}{GSS Data Explorer page for \code{away2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no answer \tab don't know \tab not available in this release \tab Total \cr
   1975 \tab 1479 \tab 11 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1495 \tab 4 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1522 \tab 8 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1513 \tab 19 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1457 \tab 11 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1853 \tab 7 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1585 \tab 14 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1459 \tab 14 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1527 \tab 7 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1465 \tab 5 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1806 \tab 13 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1480 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1533 \tab 4 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 14 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1498 \tab 19 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1589 \tab 17 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2963 \tab 29 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2888 \tab 13 \tab 3 \tab - \tab - \tab 2904 \cr
   1998 \tab 2816 \tab 16 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2812 \tab 5 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2781 \tab 31 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4471 \tab 39 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1992 \tab 29 \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 2017 \tab 25 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1955 \tab 19 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2495 \tab 42 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 2825 \tab 42 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2321 \tab 27 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 59720 \tab 485 \tab 7 \tab 1 \tab 7576 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away1}},
\code{\link{gender1}},
\code{\link{gender2}},
\code{\link{mar1}},
\code{\link{mar2}},
\code{\link{old1}},
\code{\link{old2}},
\code{\link{relate2}},
\code{\link{where1}}
}
\concept{Household Members and Structure}
\keyword{variable}
