% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_584.R
\name{away4}
\alias{away4}
\title{Is 4th person staying somewhere else now?}
\source{
General Social Survey https://gss.norc.org
}
\description{
away4
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
5. Is the fourth person staying somewhere else right now?
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5354/vshow}{GSS Data Explorer page for \code{away4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no answer \tab don't know \tab not available in this release \tab Total \cr
   1975 \tab 1474 \tab 16 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1496 \tab 3 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1518 \tab 12 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1521 \tab 11 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1462 \tab 6 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1854 \tab 6 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1581 \tab 18 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1462 \tab 11 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1517 \tab 17 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1465 \tab 5 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1801 \tab 18 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1528 \tab 9 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1367 \tab 5 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1510 \tab 6 \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab 1594 \tab 12 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2970 \tab 22 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2899 \tab 4 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab 2824 \tab 8 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2812 \tab 5 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2790 \tab 21 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 4487 \tab 22 \tab - \tab 1 \tab - \tab 4510 \cr
   2008 \tab 2009 \tab 12 \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 2026 \tab 16 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1967 \tab 7 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2522 \tab 16 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2850 \tab 16 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 2336 \tab 10 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 59888 \tab 314 \tab 9 \tab 2 \tab 7576 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away5}},
\code{\link{gender5}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{old5}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{where4}},
\code{\link{where5}}
}
\concept{Household Members and Structure}
\keyword{variable}
