% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_585.R
\name{away6}
\alias{away6}
\title{Is 6th person staying somewhere else now?}
\source{
General Social Survey https://gss.norc.org
}
\description{
away6
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
5. Is the sixth person staying somewhere else right now?
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5366/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab don't know \tab no answer \tab not available in this year \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1489 \tab 1 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1529 \tab 1 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1529 \tab 3 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1858 \tab 2 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1594 \tab 5 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1471 \tab 2 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1533 \tab 1 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1469 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1816 \tab 3 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1536 \tab 1 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1370 \tab 2 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1604 \tab 2 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2990 \tab 2 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2902 \tab 2 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2809 \tab 3 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4508 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2019 \tab 4 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2041 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1973 \tab 1 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2537 \tab 1 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2866 \tab 1 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2346 \tab 2 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab 3543 \tab 1 \tab - \tab - \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 68312 \tab 43 \tab 1 \tab 2 \tab 4032 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}

Other Family: 
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}

Other Household: 
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}
}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
