% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_15.R
\name{babies}
\alias{babies}
\title{Household members less than 6 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
babies
}
\details{
Question Household Size and Composition (see Appendix D:  Recodes, for information about these variables).  A. NUMBER OF MEMBERS UNDER 6 YEARS OLD:
As of 2024, BABIES has been recoded to only account for 0 to 2+ children younger than 6 in the household. This has been retroactively applied to all previous years of BABIES. The original BABIES containing data from 1972 to 2022 has been renamed BABIES_7222.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0
\item \verb{[1]} 1
\item \verb{[2]} 2 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/98/vshow}{GSS Data Explorer page for \code{babies}}.

Counts by year:\tabular{llllllll}{
   year \tab 0 \tab 1 \tab 2 or more \tab no answer \tab skipped on web \tab iap \tab Total \cr
   1972 \tab 1215 \tab 228 \tab 170 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1150 \tab 216 \tab 138 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1127 \tab 227 \tab 128 \tab 2 \tab - \tab - \tab 1484 \cr
   1975 \tab 1167 \tab 228 \tab 95 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1160 \tab 224 \tab 113 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab 1248 \tab 197 \tab 76 \tab 9 \tab - \tab - \tab 1530 \cr
   1978 \tab 1232 \tab 195 \tab 91 \tab 14 \tab - \tab - \tab 1532 \cr
   1980 \tab 1179 \tab 200 \tab 78 \tab 11 \tab - \tab - \tab 1468 \cr
   1982 \tab 1507 \tab 240 \tab 108 \tab 5 \tab - \tab - \tab 1860 \cr
   1983 \tab 1275 \tab 212 \tab 107 \tab 5 \tab - \tab - \tab 1599 \cr
   1984 \tab 1203 \tab 176 \tab 89 \tab 5 \tab - \tab - \tab 1473 \cr
   1985 \tab 1244 \tab 200 \tab 85 \tab 5 \tab - \tab - \tab 1534 \cr
   1986 \tab 1189 \tab 194 \tab 82 \tab 5 \tab - \tab - \tab 1470 \cr
   1987 \tab 1509 \tab 215 \tab 93 \tab 2 \tab - \tab - \tab 1819 \cr
   1988 \tab 1200 \tab 202 \tab 68 \tab 11 \tab - \tab - \tab 1481 \cr
   1989 \tab 1233 \tab 192 \tab 91 \tab 21 \tab - \tab - \tab 1537 \cr
   1990 \tab 1134 \tab 157 \tab 71 \tab 10 \tab - \tab - \tab 1372 \cr
   1991 \tab 1231 \tab 183 \tab 97 \tab 6 \tab - \tab - \tab 1517 \cr
   1993 \tab 1321 \tab 198 \tab 84 \tab 3 \tab - \tab - \tab 1606 \cr
   1994 \tab 2485 \tab 369 \tab 135 \tab 3 \tab - \tab - \tab 2992 \cr
   1996 \tab 2408 \tab 340 \tab 156 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2401 \tab 308 \tab 118 \tab 5 \tab - \tab - \tab 2832 \cr
   2000 \tab 2322 \tab 315 \tab 133 \tab 47 \tab - \tab - \tab 2817 \cr
   2002 \tab 2379 \tab 259 \tab 98 \tab 29 \tab - \tab - \tab 2765 \cr
   2004 \tab 2439 \tab 241 \tab 112 \tab 20 \tab - \tab - \tab 2812 \cr
   2006 \tab 3850 \tab 421 \tab 201 \tab 38 \tab - \tab - \tab 4510 \cr
   2008 \tab 1697 \tab 187 \tab 110 \tab 29 \tab - \tab - \tab 2023 \cr
   2010 \tab 1759 \tab 201 \tab 76 \tab 8 \tab - \tab - \tab 2044 \cr
   2012 \tab 1686 \tab 180 \tab 88 \tab 20 \tab - \tab - \tab 1974 \cr
   2014 \tab 2275 \tab 177 \tab 82 \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab 2550 \tab 192 \tab 101 \tab 24 \tab - \tab - \tab 2867 \cr
   2018 \tab 2073 \tab 185 \tab 61 \tab 29 \tab - \tab - \tab 2348 \cr
   2021 \tab 3641 \tab 254 \tab 134 \tab - \tab 3 \tab - \tab 4032 \cr
   2022 \tab 1577 \tab 130 \tab 61 \tab 8 \tab 4 \tab 1764 \tab 3544 \cr
   2024 \tab 1475 \tab 139 \tab 62 \tab 1 \tab 2 \tab 1630 \tab 3309 \cr
   Total \tab 60541 \tab 7782 \tab 3592 \tab 381 \tab 9 \tab 3394 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{adults}},
\code{\link{earnrs}},
\code{\link{granborn}},
\code{\link{hompop}},
\code{\link{income}},
\code{\link{preteen}},
\code{\link{rincome}},
\code{\link{teens}},
\code{\link{unrelat}}

Other Household: 
\code{\link{adults}},
\code{\link{earnrs}},
\code{\link{hompop}},
\code{\link{preteen}},
\code{\link{teens}},
\code{\link{unrelat}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
