% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_184.R
\name{badbrks}
\alias{badbrks}
\title{Most of my problems are due to bad breaks.}
\source{
General Social Survey https://gss.norc.org
}
\description{
badbrks
}
\details{
Question The following are some statements that people have made. For each one, please tell me if you strongly agree, agree, disagree, or strongly disagree with the statement. (NOTE: INTERVIEWERS SHOULD CODE "neither agree or disagree/it depends" ONLY WHEN VOLUNTEERED.)
B. Most of my problems are due to bad breaks.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree or disagree-it depends
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1726/vshow}{GSS Data Explorer page for \code{badbrks}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree or disagree-it depends \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 43 \tab 210 \tab 134 \tab 782 \tab 277 \tab 11 \tab 3 \tab 2904 \cr
   Total \tab 1444 \tab 43 \tab 210 \tab 134 \tab 782 \tab 277 \tab 11 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angry}},
\code{\link{embarrss}},
\code{\link{litcntrl}},
\code{\link{mostluck}},
\code{\link{noemote}},
\code{\link{noplan}},
\code{\link{proud}},
\code{\link{showangr}},
\code{\link{showfeel}}
}
\concept{Emotions}
\keyword{variable}
