% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{badchar}
\alias{badchar}
\title{Situation caused by own bad character}
\source{
General Social Survey https://gss.norc.org
}
\description{
badchar
}
\details{
Question A. Now, I want to ask you some specific questions about (John/Mary). In your opinion, how likely is it that (John/Mary)'s situation MIGHT be caused by (his/her) own bad character--very likely, somewhat likely, not very likely, or not at all likely?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2921/vshow}{GSS Data Explorer page for \code{badchar}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 77 \tab 334 \tab 463 \tab 470 \tab 38 \tab 11 \tab 2765 \cr
   Total \tab 1372 \tab 77 \tab 334 \tab 463 \tab 470 \tab 38 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldprob}},
\code{\link{chldvig}},
\code{\link{stressfl}}
}
\concept{Mental Health II}
\keyword{variable}
