% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{ballotformwt}
\alias{ballotformwt}
\title{Ballotformwt}
\source{
General Social Survey https://gss.norc.org
}
\description{
ballotformwt
}
\details{
Question Person weight for years 2002, 2010, 2012, 2016, and 2018 due to an assignment error of ballot and form for these years. For more information please see Methodological Report #134 at https://gss.norc.org/us/en/gss/get-documentation/methodological-reports.html.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2010, 2012, 2016, 2018 \cr
}
}

\keyword{variable}
