% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{ballotformwtnr}
\alias{ballotformwtnr}
\title{Ballotformwtnr}
\source{
General Social Survey https://gss.norc.org
}
\description{
ballotformwtnr
}
\details{
Question Non-response person weight for years 2002, 2010, 2012, 2016, and 2018 due to an assignment error of ballot and form for these years. For more information please see Methodological Report #134 at https://gss.norc.org/us/en/gss/get-documentation/methodological-reports.html.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2012, 2016, 2018 \cr
}
}

\keyword{variable}
