% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{balneg}
\alias{balneg}
\title{Sci research is strongly in favor of harmful results}
\source{
General Social Survey https://gss.norc.org
}
\description{
balneg
}
\details{
Question Now for another type of question. People have frequently noted that scientific research has produced benefits and harmful results. Would you say that, on balance, the benefits of scientific research have outweighed the harmful results, or have the harmful results of scientific research been greater than its benefits?
B. Would you say that the balance has been strongly in favor of the harmful results, or only slightly?
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor
\item \code{2} slightly in favor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3327/vshow}{GSS Data Explorer page for \code{balneg}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab strongly in favor \tab slightly in favor \tab don't know \tab no answer \tab Total \cr
   2006 \tab 4380 \tab 39 \tab 87 \tab 3 \tab 1 \tab 4510 \cr
   2008 \tab 1869 \tab 40 \tab 105 \tab 3 \tab 6 \tab 2023 \cr
   2010 \tab 1982 \tab 15 \tab 40 \tab - \tab 7 \tab 2044 \cr
   2012 \tab 1877 \tab 26 \tab 60 \tab 1 \tab 10 \tab 1974 \cr
   2014 \tab 2415 \tab 28 \tab 92 \tab 2 \tab 1 \tab 2538 \cr
   2016 \tab 2747 \tab 28 \tab 83 \tab 1 \tab 8 \tab 2867 \cr
   2018 \tab 2222 \tab 25 \tab 89 \tab 3 \tab 9 \tab 2348 \cr
   2021 \tab 3947 \tab 21 \tab 42 \tab - \tab 22 \tab 4032 \cr
   Total \tab 21439 \tab 222 \tab 598 \tab 13 \tab 64 \tab 22336 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{odds1}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
