% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{balwkfam}
\alias{balwkfam}
\title{How successful balancing work & family}
\source{
General Social Survey https://gss.norc.org
}
\description{
balwkfam
}
\details{
Question How successful do you feel at balancing your paid work and your family life? Do you feel...
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all successful
\item \verb{[2]} not very successful
\item \verb{[3]} somewhat successful
\item \verb{[4]} very successful
\item \verb{[5]} completely successful
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1666/vshow}{GSS Data Explorer page for \code{balwkfam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all successful \tab not very successful \tab somewhat successful \tab very successful \tab completely successful \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1910 \tab 22 \tab 112 \tab 479 \tab 310 \tab 50 \tab 11 \tab 10 \tab 2904 \cr
   Total \tab 1910 \tab 22 \tab 112 \tab 479 \tab 310 \tab 50 \tab 11 \tab 10 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{earnmore}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}
}
\concept{Gender}
\keyword{variable}
