% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{baseofop}
\alias{baseofop}
\title{You have base of operations}
\source{
General Social Survey https://gss.norc.org
}
\description{
baseofop
}
\details{
Question Do you have any kind of base of operations at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1374/vshow}{GSS Data Explorer page for \code{baseofop}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 1422 \tab 80 \tab 6 \tab 9 \tab 1517 \cr
   Total \tab 1422 \tab 80 \tab 6 \tab 9 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{splocnum}},
\code{\link{training}},
\code{\link{wherewrk}}
}
\concept{Work Organizations}
\keyword{variable}
