% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{batch}
\alias{batch}
\title{Sample batch}
\source{
General Social Survey https://gss.norc.org
}
\description{
batch
}
\details{
Question Sample batch
}
\section{Values}{

\itemize{
\item \verb{[1]} batch 1
\item \verb{[2]} batch 2
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8038/vshow}{GSS Data Explorer page for \code{batch}}.

Counts by year:\tabular{llllll}{
   year \tab not available in this year \tab batch 1 \tab batch 2 \tab iap \tab Total \cr
   2022 \tab - \tab 1211 \tab 793 \tab 1540 \tab 3544 \cr
   Total \tab 0 \tab 1211 \tab 793 \tab 1540 \tab 3544 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{cshutyp12}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{subsamprate}},
\code{\link{version}},
\code{\link{vpsu}},
\code{\link{vstrat}},
\code{\link{wtss}},
\code{\link{wtssall}},
\code{\link{wtssnr}}
}
\concept{Sampling and Weights}
\keyword{variable}
