% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{befair}
\alias{befair}
\title{How often do you think people take advantage}
\source{
General Social Survey https://gss.norc.org
}
\description{
befair
}
\details{
Question How often do you think that people would try to take advantage of you if they got the chance and how often would they try to be fair?
}
\section{Values}{

\itemize{
\item \verb{[1]} try to take advantage almost all of the time
\item \verb{[2]} try to take advantage most of the time
\item \verb{[3]} try to be fair most of the time
\item \verb{[4]} try to be fair almost all of the time
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2004, 2014 \cr
   B/C/- \tab 2018, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4685/vshow}{GSS Data Explorer page for \code{befair}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab try to take advantage almost all of the time \tab try to take advantage most of the time \tab try to be fair most of the time \tab try to be fair almost all of the time \tab can't choose \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1998 \tab 1548 \tab 123 \tab 321 \tab 581 \tab 166 \tab 64 \tab 29 \tab - \tab - \tab 2832 \cr
   2004 \tab 1340 \tab 115 \tab 368 \tab 845 \tab 125 \tab 12 \tab 7 \tab - \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 104 \tab 378 \tab 633 \tab 71 \tab 76 \tab 2 \tab - \tab - \tab 2538 \cr
   2018 \tab 1175 \tab 81 \tab 240 \tab 654 \tab 156 \tab 37 \tab 5 \tab - \tab - \tab 2348 \cr
   2024 \tab 1667 \tab 185 \tab 462 \tab 689 \tab 115 \tab 159 \tab 6 \tab - \tab 26 \tab 3309 \cr
   Total \tab 7004 \tab 608 \tab 1769 \tab 3402 \tab 633 \tab 348 \tab 49 \tab 0 \tab 26 \tab 13839 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{bible1}},
\code{\link{cantrust}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conbiz}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{hapunhap}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{madenkid}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relcand}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}

Other People: 
\code{\link{advantge}},
\code{\link{cantrust}},
\code{\link{fair}},
\code{\link{fair5}},
\code{\link{helpful}},
\code{\link{trust}},
\code{\link{trust5}},
\code{\link{trusting}},
\code{\link{trustpeo}},
\code{\link{trusty}},
\code{\link{twoclass}},
\code{\link{wantbest}}
}
\concept{ISSP Religion}
\concept{People}
\keyword{variable}
