% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{belikeus}
\alias{belikeus}
\title{Agree better if people were more like americans}
\source{
General Social Survey https://gss.norc.org
}
\description{
belikeus
}
\details{
Question How much do you agree or disagree with the following statements?
C. The world would be a better place if people from other countries were more like the Americans
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4836/vshow}{GSS Data Explorer page for \code{belikeus}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 193 \tab 330 \tab 439 \tab 281 \tab 56 \tab 52 \tab 16 \tab 2904 \cr
   2004 \tab 1596 \tab 190 \tab 296 \tab 419 \tab 224 \tab 68 \tab 18 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 86 \tab 283 \tab 417 \tab 340 \tab 86 \tab 58 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 469 \tab 909 \tab 1275 \tab 845 \tab 210 \tab 128 \tab 21 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}}
}
\concept{ISSP National Identity}
\keyword{variable}
