% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_185.R
\name{beplesnt}
\alias{beplesnt}
\title{I try to be pleasant so not to upset oth.}
\source{
General Social Survey https://gss.norc.org
}
\description{
beplesnt
}
\details{
Question The following are some statements that people have made. For each one, please tell me if you strongly agree, agree, disagree, or strongly disagree with the statement. (NOTE: INTERVIEWERS SHOULD CODE "neither agree or disagree/it depends" ONLY WHEN VOLUNTEERED.)
I. I try to be pleasant so that others won't get upset.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree or disagree-it depends
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1733/vshow}{GSS Data Explorer page for \code{beplesnt}}.
}

\seealso{
Other Emotions: 
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{madat1}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{whnangry}}
}
\concept{Emotions}
\keyword{variable}
