% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{betrlang}
\alias{betrlang}
\title{Which language r speaks more fluent}
\source{
General Social Survey https://gss.norc.org
}
\description{
betrlang
}
\details{
Question B. Which language do you speak more fluent?
}
\section{Values}{

\itemize{
\item \verb{[1]} language 1
\item \verb{[2]} language 2
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2302/vshow}{GSS Data Explorer page for \code{betrlang}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab language 1 \tab language 2 \tab no answer \tab don't know \tab skipped on web \tab not available in this year \tab Total \cr
   2006 \tab 3837 \tab 635 \tab 36 \tab 2 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1471 \tab 515 \tab 32 \tab - \tab 5 \tab - \tab - \tab 2023 \cr
   2010 \tab 1508 \tab 500 \tab 33 \tab - \tab 3 \tab - \tab - \tab 2044 \cr
   2012 \tab 1401 \tab 537 \tab 33 \tab 1 \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab 1790 \tab 699 \tab 43 \tab - \tab 6 \tab - \tab - \tab 2538 \cr
   2016 \tab 2045 \tab 749 \tab 62 \tab 8 \tab 3 \tab - \tab - \tab 2867 \cr
   2018 \tab 1623 \tab 690 \tab 31 \tab 1 \tab 3 \tab - \tab - \tab 2348 \cr
   2021 \tab 2972 \tab 1004 \tab 46 \tab 1 \tab 2 \tab 7 \tab - \tab 4032 \cr
   2022 \tab 2605 \tab 887 \tab 49 \tab - \tab 1 \tab 2 \tab - \tab 3544 \cr
   Total \tab 19252 \tab 6216 \tab 365 \tab 13 \tab 25 \tab 9 \tab 0 \tab 25880 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blkwrk}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{ethchng}},
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{knwwht}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{leastcom}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{uselang}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{uswht}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}

Other Language: 
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{english}},
\code{\link{engoff1}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{uselang}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
