% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{bettrlfe}
\alias{bettrlfe}
\title{Science makes our lives better}
\source{
General Social Survey https://gss.norc.org
}
\description{
bettrlfe
}
\details{
Question I'm going to read to you some statements like those you might find in a newspaper or magazine article. For each statement, please tell me if you strongly agree, agree, disagree, or strongly disagree.
C. Science and technology are making our lives, healthier, easier, and more comfortable.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3550/vshow}{GSS Data Explorer page for \code{bettrlfe}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1299 \tab 225 \tab 772 \tab 201 \tab 19 \tab 20 \tab 2 \tab 2538 \cr
   Total \tab 1299 \tab 225 \tab 772 \tab 201 \tab 19 \tab 20 \tab 2 \tab 2538 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
