% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{bible1}
\alias{bible1}
\title{Describe your feelings about the bible}
\source{
General Social Survey https://gss.norc.org
}
\description{
bible1
}
\details{
Question Which of these statements comes closest to describing your feelings about the Bible?
}
\section{Values}{

\itemize{
\item \verb{[1]} word of god, take literally
\item \verb{[2]} god-inspired, shouldn't take literally
\item \verb{[3]} book of fables, by man
\item \verb{[4]} this does not apply to me
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4656/vshow}{GSS Data Explorer page for \code{bible1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab word of god, take literally \tab god-inspired, shouldn't take literally \tab book of fables, by man \tab this does not apply to me \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 430 \tab 639 \tab 188 \tab 27 \tab 50 \tab 25 \tab 1517 \cr
   1998 \tab 1548 \tab 358 \tab 585 \tab 207 \tab 40 \tab 56 \tab 38 \tab 2832 \cr
   Total \tab 1706 \tab 788 \tab 1224 \tab 395 \tab 67 \tab 106 \tab 63 \tab 4349 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
