% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{biblenv}
\alias{biblenv}
\title{Feelings about the bible (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
biblenv
}
\details{
Question Which of these statements comes closest to describing your feelings about the Bible?
A.) The Bible is the actual word of God and is to be taken literally, word for word.
B.) The Bible is the inspired word of God but not everything in it should be taken
literally, word for word.
C.) The Bible is an ancient book of fables, legends, history, and moral precepts recorded by man.
}
\section{Values}{

\itemize{
\item \code{1} word of god
\item \code{2} inspired word
\item \code{3} ancient book
\item \code{4} other (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7617/vshow}{GSS Data Explorer page for \code{biblenv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab word of god \tab inspired word \tab ancient book \tab other (phone mode only: volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2067 \tab 415 \tab 929 \tab 587 \tab 4 \tab 2 \tab 1 \tab 27 \tab 4032 \cr
   2022 \tab 2688 \tab 177 \tab 370 \tab 302 \tab - \tab - \tab - \tab 7 \tab 3544 \cr
   Total \tab 4755 \tab 592 \tab 1299 \tab 889 \tab 4 \tab 2 \tab 1 \tab 34 \tab 7576 \cr
}
}

\keyword{variable}
