% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{bibley}
\alias{bibley}
\title{Feelings about the bible:version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
bibley
}
\details{
Question B. Here are four statements about the Bible, and I'd like you to tell me which is closest to your own view. READ FIRST FOUR STATEMENTS ((a)-(d)) ONLY. CIRCLE ONLY ONE CODE.
a. The Bible is God's Word and all it says is true......1
b.The Bible was written by men inspired by God, but it contains some human errors ......2
c. The Bible is a good book because it was written by wise men, but God had nothing to do with it ......3
d. The Bible was written by men who lived so long ago that it is worth very little today......4
}
\section{Values}{

\itemize{
\item \code{1} god's word
\item \code{2} inspired by god
\item \code{3} written by man
\item \code{4} not worth much
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
   1985 \tab - \tab full \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/365/vshow}{GSS Data Explorer page for \code{bibley}}.
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}

Other Religion: 
\code{\link{bible}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}

Other Split Ballots: 
\code{\link{bible}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}}
}
\concept{Core}
\concept{Religion}
\concept{Split Ballots}
\keyword{variable}
