% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_545.R
\name{big5a2}
\alias{big5a2}
\title{R see oneself as someone who is sociable}
\source{
General Social Survey https://gss.norc.org
}
\description{
big5a2
}
\details{
Question To what extent do you agree or disagree with the following statements? I see myself as someone who…
A. …is outgoing, sociable
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4979/vshow}{GSS Data Explorer page for \code{big5a2}}.
}

\seealso{
Other Big 5: 
\code{\link{big5b2}},
\code{\link{big5c2}},
\code{\link{big5d2}},
\code{\link{big5e1}},
\code{\link{big5e2}}

Other Feelings: 
\code{\link{big5b2}},
\code{\link{big5c2}},
\code{\link{big5d2}},
\code{\link{big5e1}},
\code{\link{big5e2}}
}
\concept{Big 5}
\concept{Feelings}
\keyword{variable}
