% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{big5c1}
\alias{big5c1}
\title{R see oneself as someone who does a thorough job}
\source{
General Social Survey https://gss.norc.org
}
\description{
big5c1
}
\details{
Question To what extent do you agree or disagree with the following statements? I see myself as someone who…
C. …does a thorough job
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4976/vshow}{GSS Data Explorer page for \code{big5c1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 743 \tab 719 \tab 43 \tab 9 \tab 2 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 743 \tab 719 \tab 43 \tab 9 \tab 2 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Big 5: 
\code{\link{big5a1}},
\code{\link{big5b1}},
\code{\link{big5d1}}
}
\concept{Big 5}
\keyword{variable}
