% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{big5d1}
\alias{big5d1}
\title{R see oneself as someone who is relaxed}
\source{
General Social Survey https://gss.norc.org
}
\description{
big5d1
}
\details{
Question To what extent do you agree or disagree with the following statements? I see myself as someone who…
D. …is relaxed, handles stress well
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4977/vshow}{GSS Data Explorer page for \code{big5d1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 305 \tab 692 \tab 247 \tab 219 \tab 52 \tab 1 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 305 \tab 692 \tab 247 \tab 219 \tab 52 \tab 1 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Big 5: 
\code{\link{big5a1}},
\code{\link{big5b1}},
\code{\link{big5c1}}
}
\concept{Big 5}
\keyword{variable}
