% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_545.R
\name{big5e1}
\alias{big5e1}
\title{R see oneself as someone who has an active imagination}
\source{
General Social Survey https://gss.norc.org
}
\description{
big5e1
}
\details{
Question To what extent do you agree or disagree with the following statements? I see myself as someone who…
E. …has an active imagination
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4978/vshow}{GSS Data Explorer page for \code{big5e1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 463 \tab 688 \tab 204 \tab 140 \tab 15 \tab 6 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 463 \tab 688 \tab 204 \tab 140 \tab 15 \tab 6 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Big 5: 
\code{\link{big5a2}},
\code{\link{big5b2}},
\code{\link{big5c2}},
\code{\link{big5d2}},
\code{\link{big5e2}}
}
\concept{Big 5}
\keyword{variable}
