% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_545.R
\name{big5e2}
\alias{big5e2}
\title{R see oneself as someone who has few artistic interests}
\source{
General Social Survey https://gss.norc.org
}
\description{
big5e2
}
\details{
Question To what extent do you agree or disagree with the following statements? I see myself as someone who…
E. …has few artistic interests
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4983/vshow}{GSS Data Explorer page for \code{big5e2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 83 \tab 423 \tab 218 \tab 547 \tab 241 \tab 4 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 83 \tab 423 \tab 218 \tab 547 \tab 241 \tab 4 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Big 5: 
\code{\link{big5a2}},
\code{\link{big5b2}},
\code{\link{big5c2}},
\code{\link{big5d2}},
\code{\link{big5e1}}
}
\concept{Big 5}
\keyword{variable}
