% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{bigbang}
\alias{bigbang}
\title{Sci knowledge: the universe began with a huge explosion}
\source{
General Social Survey https://gss.norc.org
}
\description{
bigbang
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don't know or aren't sure, just tell me so,  and we will skip to the next question. Remember true, false, or don't know.
G. The universe began with a huge explosion. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3340/vshow}{GSS Data Explorer page for \code{bigbang}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 590 \tab 633 \tab 639 \tab 2 \tab 4510 \cr
   2008 \tab 518 \tab 482 \tab 606 \tab 402 \tab 15 \tab 2023 \cr
   2010 \tab 1103 \tab 345 \tab 316 \tab 277 \tab 3 \tab 2044 \cr
   2012 \tab 1457 \tab 195 \tab 163 \tab 156 \tab 3 \tab 1974 \cr
   2014 \tab 1299 \tab 510 \tab 437 \tab 288 \tab 4 \tab 2538 \cr
   2016 \tab 2410 \tab 179 \tab 161 \tab 111 \tab 6 \tab 2867 \cr
   2018 \tab 1944 \tab 153 \tab 146 \tab 103 \tab 2 \tab 2348 \cr
   Total \tab 11377 \tab 2454 \tab 2462 \tab 1976 \tab 35 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{hotcore}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{radioact}},
\code{\link{viruses}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
