% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{bigbang1}
\alias{bigbang1}
\title{Universe began with a big explosion: true or false}
\source{
General Social Survey https://gss.norc.org
}
\description{
bigbang1
}
\details{
Question According to astronomers, the universe began with a huge explosion. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/- \tab 2016, 2018 \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3548/vshow}{GSS Data Explorer page for \code{bigbang1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 287 \tab 85 \tab 109 \tab 4 \tab 1974 \cr
   2016 \tab 2388 \tab 320 \tab 80 \tab 75 \tab 4 \tab 2867 \cr
   2018 \tab 1958 \tab 255 \tab 83 \tab 52 \tab - \tab 2348 \cr
   Total \tab 5835 \tab 862 \tab 248 \tab 236 \tab 8 \tab 7189 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bigbang}},
\code{\link{bigbang2}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{hotcore}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{radioact}},
\code{\link{viruses}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
