% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{bilinged}
\alias{bilinged}
\title{Support or oppose bilingual education}
\source{
General Social Survey https://gss.norc.org
}
\description{
bilinged
}
\details{
Question How do you feel about bilingual education? Are you strongly in favor of it, somewhat in favor of it, somewhat opposed to it, or strongly opposed to it?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} somewhat favor
\item \verb{[3]} somewhat opposed
\item \verb{[4]} strongly opposed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1607/vshow}{GSS Data Explorer page for \code{bilinged}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab somewhat favor \tab somewhat opposed \tab strongly opposed \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1518 \tab 388 \tab 560 \tab 240 \tab 212 \tab 67 \tab 7 \tab - \tab 2992 \cr
   Total \tab 1518 \tab 388 \tab 560 \tab 240 \tab 212 \tab 67 \tab 7 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{econfutr}},
\code{\link{econpast}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hspasn10}},
\code{\link{hspgovt}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{jobaff}},
\code{\link{letin}},
\code{\link{meltpot}},
\code{\link{obrespct}},
\code{\link{owneth}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{united}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Education: 
\code{\link{actgreek}},
\code{\link{actschl}},
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{cathsch}},
\code{\link{chldeduc}},
\code{\link{chrissch}},
\code{\link{classics}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{coldeg1}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{colop}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{datesch}},
\code{\link{degree}},
\code{\link{denomsch}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{educ}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{educemp}},
\code{\link{educop}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{govedop}},
\code{\link{homesch}},
\code{\link{hsbasics}},
\code{\link{hsbio}},
\code{\link{hscaring}},
\code{\link{hschem}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsmath}},
\code{\link{hsorder}},
\code{\link{hsphys}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inteduc}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{majorcol}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memschl}},
\code{\link{natdeg}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{opeduc}},
\code{\link{oppared}},
\code{\link{othsch}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{relsch}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{schleth}},
\code{\link{schlhead}},
\code{\link{schoolng}},
\code{\link{scimath}},
\code{\link{sexeduc}},
\code{\link{solgreek}},
\code{\link{solschl}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{speduc}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{teacheth}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{trstprof}},
\code{\link{undoccol}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{whoteach}}

Other Language: 
\code{\link{betrlang}},
\code{\link{engballt}},
\code{\link{english}},
\code{\link{engoff1}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{uselang}}
}
\concept{Education}
\concept{Language}
\concept{Multiculturalism}
\keyword{variable}
