% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{biosci}
\alias{biosci}
\title{How scientific: biology}
\source{
General Social Survey https://gss.norc.org
}
\description{
biosci
}
\details{
Question Please look at Card B20. How scientific are each of the following fields? If you have not heard of a particular field, just say you haven't heard of it
E. Biology. Is biology very scientific, pretty scientific, not too scientific, or not scientific at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very scientific
\item \verb{[2]} pretty scientific
\item \verb{[3]} not too scientific
\item \verb{[4]} not scientific at all
\item \verb{[5]} haven't heard of it (volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3410/vshow}{GSS Data Explorer page for \code{biosci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab haven't heard of it (volunteered) \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1300 \tab 447 \tab 39 \tab 14 \tab 14 \tab 46 \tab 4 \tab 4510 \cr
   2012 \tab 1477 \tab 338 \tab 106 \tab 21 \tab 6 \tab 3 \tab 19 \tab 4 \tab 1974 \cr
   Total \tab 4123 \tab 1638 \tab 553 \tab 60 \tab 20 \tab 17 \tab 65 \tab 8 \tab 6484 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{medsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
