% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{bizbstgw}
\alias{bizbstgw}
\title{Best policy vs narrow interest: business leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizbstgw
}
\details{
Question Please look at Card B5. When making policy recommendations about global warming, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole versus what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
C. Business Leaders
}
\section{Values}{

\itemize{
\item \verb{[1]} what is best for the country
\item \verb{[5]} own narrow interests
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3364/vshow}{GSS Data Explorer page for \code{bizbstgw}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 51 \tab 39 \tab 212 \tab 259 \tab 320 \tab 45 \tab 2 \tab 4510 \cr
   2010 \tab 1569 \tab 25 \tab 25 \tab 118 \tab 126 \tab 156 \tab 20 \tab 5 \tab 2044 \cr
   Total \tab 5151 \tab 76 \tab 64 \tab 330 \tab 385 \tab 476 \tab 65 \tab 7 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizinfgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
