% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{bizbstnk}
\alias{bizbstnk}
\title{Business leaders best for country}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizbstnk
}
\details{
Question Please look at Card E16. When making policy recommendations about nuclear power, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests? FI INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
C. Business leaders
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3497/vshow}{GSS Data Explorer page for \code{bizbstnk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1578 \tab 23 \tab 40 \tab 124 \tab 146 \tab 102 \tab 24 \tab 7 \tab 2044 \cr
   Total \tab 1578 \tab 23 \tab 40 \tab 124 \tab 146 \tab 102 \tab 24 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{nukeelec}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
