% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_374.R
\name{bizbsttx}
\alias{bizbsttx}
\title{Best policy vs narrow interest on fed inc tax: business leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizbsttx
}
\details{
Question Please look at Card B14. When making policy recommendations about federal income taxes, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
B. Business leaders
}
\section{Values}{

\itemize{
\item \verb{[1]} what is best for the country
\item \verb{[5]} own narrow interests
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3385/vshow}{GSS Data Explorer page for \code{bizbsttx}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3574 \tab 35 \tab 74 \tab 229 \tab 283 \tab 268 \tab 47 \tab - \tab 4510 \cr
   2010 \tab 1578 \tab 22 \tab 50 \tab 106 \tab 144 \tab 121 \tab 19 \tab 4 \tab 2044 \cr
   Total \tab 5152 \tab 57 \tab 124 \tab 335 \tab 427 \tab 389 \tab 66 \tab 4 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizinftx}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{polbsttx}},
\code{\link{polinftx}},
\code{\link{txpol}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
