% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{bizgross}
\alias{bizgross}
\title{Business revenue last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizgross
}
\details{
Question In which of these groups did the revenues from this business fall for last year
}
\section{Values}{

\itemize{
\item \verb{[1]} ( $15,000
\item \verb{[2]} $15,001-$25,000
\item \verb{[3]} $25,001-$30,000
\item \verb{[4]} $30,001-$35,000
\item \verb{[5]} $35,001-$50,000
\item \verb{[6]} $50,001-$60,000
\item \verb{[7]} $60,001-$75,000
\item \verb{[8]} $75,001-$100,000
\item \verb{[9]} ) $100,000
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3819/vshow}{GSS Data Explorer page for \code{bizgross}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab ( $15,000 \tab $15,001-$25,000 \tab $25,001-$30,000 \tab $30,001-$35,000 \tab $35,001-$50,000 \tab $50,001-$60,000 \tab $60,001-$75,000 \tab $75,001-$100,000 \tab ) $100,000 \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 1773 \tab 104 \tab 16 \tab 11 \tab 7 \tab 17 \tab 9 \tab 7 \tab 7 \tab 48 \tab 12 \tab 12 \tab - \tab 2023 \cr
   Total \tab 1773 \tab 104 \tab 16 \tab 11 \tab 7 \tab 17 \tab 9 \tab 7 \tab 7 \tab 48 \tab 12 \tab 12 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{numown}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{startbiz}},
\code{\link{whybiz}},
\code{\link{yearbiz}}

Other Business: 
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizshare}},
\code{\link{busdecid}},
\code{\link{buspow}},
\code{\link{bustax}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{numemply}},
\code{\link{numemps}},
\code{\link{numemps5}},
\code{\link{numemps_0422}},
\code{\link{numown}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{paidemps}},
\code{\link{privent}},
\code{\link{startbiz}},
\code{\link{whybiz}},
\code{\link{yearbiz}}

Other Corporations: 
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizshare}},
\code{\link{equal2}},
\code{\link{equal6}},
\code{\link{merged}},
\code{\link{orgfin}},
\code{\link{outofbiz}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{startbiz}},
\code{\link{whybiz}},
\code{\link{yearbiz}}

Other Employment: 
\code{\link{accntsci}},
\code{\link{advances}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizshare}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compwage}},
\code{\link{curempyr}},
\code{\link{denyrais}},
\code{\link{dowell}},
\code{\link{drugtest}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empyears}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{evunemp}},
\code{\link{expernc}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findjob}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{futpromo}},
\code{\link{harjob5}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{ignorwk}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobyears}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{knowschd}},
\code{\link{laidoff}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leavejb}},
\code{\link{liedcwkr}},
\code{\link{losejb12}},
\code{\link{losejob5}},
\code{\link{lostjobs}},
\code{\link{ltrsref}},
\code{\link{marrcoun}},
\code{\link{medtreat}},
\code{\link{moredays}},
\code{\link{moretrde}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{neverwk}},
\code{\link{newjobs}},
\code{\link{numemp}},
\code{\link{numemply}},
\code{\link{numemps}},
\code{\link{numemps5}},
\code{\link{numemps_0422}},
\code{\link{numown}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{ownbiz}},
\code{\link{paidemps}},
\code{\link{partfull}},
\code{\link{paysame}},
\code{\link{physical}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{ryrsfirm}},
\code{\link{samejob}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{skiltest}},
\code{\link{slsmnshp}},
\code{\link{spwrkgvt}},
\code{\link{startbiz}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totunemp}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{whywkhme}},
\code{\link{wkbhvrs}},
\code{\link{work3yrs}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrkstat}},
\code{\link{wrktype}},
\code{\link{wrkyears}},
\code{\link{yearbiz}},
\code{\link{yearsjob}},
\code{\link{yrsfirm}}

Other Income: 
\code{\link{bizshare}},
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{earndes}},
\code{\link{earnmore}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{fairearn}},
\code{\link{famfinan}},
\code{\link{finalter}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{finrela}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{impfinan}},
\code{\link{incdef}},
\code{\link{inclstjb}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{justpay}},
\code{\link{minfood}},
\code{\link{minfour}},
\code{\link{mininc}},
\code{\link{minthree}},
\code{\link{othpay}},
\code{\link{owninc}},
\code{\link{pastpay}},
\code{\link{paychild}},
\code{\link{paychnge}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{payfam1}},
\code{\link{payhard}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{povline}},
\code{\link{realinc}},
\code{\link{realrinc}},
\code{\link{rincblls}},
\code{\link{rincom06}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{salsatfn}},
\code{\link{satfin}},
\code{\link{welfare1}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}}
}
\concept{Business}
\concept{Corporations}
\concept{Employment}
\concept{Income}
\concept{Self Employment}
\keyword{variable}
