% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{bizinfgw}
\alias{bizinfgw}
\title{Influence on global warming policy: business leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizinfgw
}
\details{
Question Please look at Card B4. How much influence should each of the following groups have in deciding what  to do about global warming?
C. Business Leaders. Would you say a great deal of influence, a fair amount, a little influence, or none at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal of influence
\item \verb{[2]} a fair amount
\item \verb{[3]} a little influence
\item \verb{[4]} none at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3361/vshow}{GSS Data Explorer page for \code{bizinfgw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal of influence \tab a fair amount \tab a little influence \tab none at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 91 \tab 211 \tab 340 \tab 241 \tab 43 \tab 2 \tab 4510 \cr
   2010 \tab 1569 \tab 37 \tab 110 \tab 200 \tab 109 \tab 15 \tab 4 \tab 2044 \cr
   Total \tab 5151 \tab 128 \tab 321 \tab 540 \tab 350 \tab 58 \tab 6 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
