% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{bizpeop}
\alias{bizpeop}
\title{Business people do more for environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizpeop
}
\details{
Question A. On balance, which of these two do you think is making more effort to look after the environment:
}
\section{Values}{

\itemize{
\item \code{1} business and industry
\item \code{2} or people in general
\item \code{3} both equally
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4784/vshow}{GSS Data Explorer page for \code{bizpeop}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}

Other Environment: 
\code{\link{amprogrn}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
