% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_344.R
\name{blessngs}
\alias{blessngs}
\title{Feel thankful for r's blessings}
\source{
General Social Survey https://gss.norc.org
}
\description{
blessngs
}
\details{
Question I feel thankful for my blessings.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never or almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3091/vshow}{GSS Data Explorer page for \code{blessngs}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never or almost never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 423 \tab 554 \tab 204 \tab 87 \tab 27 \tab 28 \tab 2 \tab 15 \tab 2812 \cr
   Total \tab 1472 \tab 423 \tab 554 \tab 204 \tab 87 \tab 27 \tab 28 \tab 2 \tab 15 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{closrgod}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{joylifts}},
\code{\link{relsprt1}},
\code{\link{relsprt2}}
}
\concept{Spirituality and Religion}
\keyword{variable}
