% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_656.R
\name{blkdsrv}
\alias{blkdsrv}
\title{Black people get less than they deserve}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkdsrv
}
\details{
Question (Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or disagree strongly with the following statements?) Over the past few years, black people have gotten less than they deserve.

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree somewhat
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7894/vshow}{GSS Data Explorer page for \code{blkdsrv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab neither agree nor disagree \tab disagree somewhat \tab disagree strongly \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 269 \tab 288 \tab 262 \tab 226 \tab 187 \tab 3 \tab 3544 \cr
   Total \tab 2309 \tab 269 \tab 288 \tab 262 \tab 226 \tab 187 \tab 3 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{blkmblty}},
\code{\link{blktry}},
\code{\link{brv5}},
\code{\link{brv5sp}}
}
\concept{GSS Next}
\keyword{variable}
