% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_175.R
\name{blkgovt}
\alias{blkgovt}
\title{Do blacks get proper govt attention}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkgovt
}
\details{
Question B. Do you think that blacks get more attention from government than they deserve?  Would you answer much more attention from government than they deserve, more attention than they deserve, about the right amount of attention, less attention than they deserve, or much less attention from government than they deserve?
}
\section{Values}{

\itemize{
\item \code{1} much more
\item \code{2} more
\item \code{3} about right
\item \code{4} less
\item \code{5} much less
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1633/vshow}{GSS Data Explorer page for \code{blkgovt}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{asngovt}},
\code{\link{econpast}},
\code{\link{hspgovt}},
\code{\link{obrespct}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Ethnicity: 
\code{\link{asngovt}},
\code{\link{hspgovt}},
\code{\link{whtgovt}}

Other Federal Government: 
\code{\link{asngovt}},
\code{\link{hspgovt}},
\code{\link{whtgovt}}

Other Race: 
\code{\link{asngovt}},
\code{\link{hspgovt}},
\code{\link{whtgovt}}
}
\concept{Blacks}
\concept{Ethnicity}
\concept{Federal Government}
\concept{Multiculturalism}
\concept{Race}
\keyword{variable}
