% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_133.R
\name{blkhouse}
\alias{blkhouse}
\title{Blacks discriminated in housing}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkhouse
}
\details{
Question How much discrimination is there that makes it hard for Hispanic Americans to buy or rent housing wherever they want?
B. How about for blacks?
}
\section{Values}{

\itemize{
\item \verb{[1]} a lot
\item \verb{[2]} some
\item \verb{[3]} only a little
\item \verb{[4]} none at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1247/vshow}{GSS Data Explorer page for \code{blkhouse}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a lot \tab some \tab only a little \tab none at all \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab - \tab 452 \tab 527 \tab 193 \tab 131 \tab 57 \tab 12 \tab - \tab 1372 \cr
   Total \tab 0 \tab 452 \tab 527 \tab 193 \tab 131 \tab 57 \tab 12 \tab 0 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Blacks: 
\code{\link{acqblack}},
\code{\link{acqfmblk}},
\code{\link{acqnhblk}},
\code{\link{acqvablk}},
\code{\link{acqwkblk}},
\code{\link{admirblk}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkgovt}},
\code{\link{blkjobs}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blksimp}},
\code{\link{blkwrk}},
\code{\link{closeblk}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{comblk}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fairblks}},
\code{\link{famblks}},
\code{\link{fareblks}},
\code{\link{feelblks}},
\code{\link{influblk}},
\code{\link{intlblks}},
\code{\link{jobaff}},
\code{\link{knwblk}},
\code{\link{liveblks}},
\code{\link{marblk}},
\code{\link{patrblks}},
\code{\link{symptblk}},
\code{\link{trtblack}},
\code{\link{usblk}},
\code{\link{violblks}},
\code{\link{wlthblks}},
\code{\link{workblks}},
\code{\link{wrkwayup}}

Other Housing: 
\code{\link{aidhouse}},
\code{\link{asnhouse}},
\code{\link{hsphouse}}
}
\concept{Blacks}
\concept{Housing}
\concept{Intergroup Relations}
\keyword{variable}
