% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{blkinflu}
\alias{blkinflu}
\title{Opinion of political influence of blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkinflu
}
\details{
Question People have different opinions about the amount of influence that various groups have in American life and politics. Do you think that blacks have far too much influence, too much influence, about the right amount of influence, too little influence, or do they far too little influence?
}
\section{Values}{

\itemize{
\item \verb{[1]} far too much
\item \verb{[2]} too much
\item \verb{[3]} right amount
\item \verb{[4]} too little
\item \verb{[5]} far too little
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1041/vshow}{GSS Data Explorer page for \code{blkinflu}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too much \tab right amount \tab too little \tab far too little \tab don't know \tab no answer \tab Total \cr
   1987 \tab 1275 \tab 7 \tab 60 \tab 268 \tab 188 \tab 16 \tab 5 \tab 1819 \cr
   Total \tab 1275 \tab 7 \tab 60 \tab 268 \tab 188 \tab 16 \tab 5 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{blkgains}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{gavepol}}
}
\concept{Socio-political participation}
\keyword{variable}
