% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{blksimp}
\alias{blksimp}
\title{Have conditions for blacks improved}
\source{
General Social Survey https://gss.norc.org
}
\description{
blksimp
}
\details{
Question In the past few years, do you think conditions for black people have improved, gotten worse, or stayed about the same?
}
\section{Values}{

\itemize{
\item \code{1} improved
\item \code{2} gotten worse
\item \code{3} about the same
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/425/vshow}{GSS Data Explorer page for \code{blksimp}}.
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}

Other Blacks: 
\code{\link{wrkwayup}}

Other Race: 
\code{\link{racfirm}},
\code{\link{racinfo}},
\code{\link{wrkwayup}}
}
\concept{Blacks}
\concept{Core}
\concept{Race}
\keyword{variable}
