% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{blkwhite}
\alias{blkwhite}
\title{R agrees right & wrong not black & white}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkwhite
}
\details{
Question Please consider the following statements and tell me whether you agree strongly, agree somewhat, disagree somewhat or disagree strongly with each statement.
B. Right and wrong are not usually a simple matter of black and white; there are many shades of gray.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1082/vshow}{GSS Data Explorer page for \code{blkwhite}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 589 \tab 624 \tab 115 \tab 102 \tab 43 \tab 8 \tab 1481 \cr
   2006 \tab 1518 \tab 1279 \tab 1149 \tab 294 \tab 203 \tab 62 \tab 5 \tab 4510 \cr
   2008 \tab - \tab 896 \tab 725 \tab 190 \tab 152 \tab 53 \tab 7 \tab 2023 \cr
   2010 \tab - \tab 890 \tab 800 \tab 163 \tab 140 \tab 45 \tab 6 \tab 2044 \cr
   Total \tab 1518 \tab 3654 \tab 3298 \tab 762 \tab 597 \tab 203 \tab 26 \tab 10058 \cr
}
}

\seealso{
Other Religion: 
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{gochurch}},
\code{\link{permoral}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
