% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_259.R
\name{blkwrk}
\alias{blkwrk}
\title{Does r know blacks from work}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkwrk
}
\details{
Question Now, I’m going to ask some questions about these people you personally know.
D. Do you know any of these Blacks from the place where you work?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2343/vshow}{GSS Data Explorer page for \code{blkwrk}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2082 \tab 494 \tab 211 \tab 1 \tab 29 \tab 2817 \cr
   Total \tab 2082 \tab 494 \tab 211 \tab 1 \tab 29 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{hispschl}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
