% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_132.R
\name{blkzone}
\alias{blkzone}
\title{Tax breaks for black areas}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkzone
}
\details{
Question Here are several things that the government in Washington might Do to deal with the problems of poverty and
unemployment among Black Americans.
I would like you to tell me if you favor or oppose them. (READ "A") - Would you say that you strongly favor it, favor it, neither favor it nor oppose it, oppose it, or strongly oppose it?
A. Giving business and industry special tax breaks for locating in largely black areas.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1233/vshow}{GSS Data Explorer page for \code{blkzone}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab 674 \tab 68 \tab 241 \tab 160 \tab 149 \tab 46 \tab 25 \tab 9 \tab 1372 \cr
   Total \tab 674 \tab 68 \tab 241 \tab 160 \tab 149 \tab 46 \tab 25 \tab 9 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{blkcol}},
\code{\link{blkschs}},
\code{\link{influblk}},
\code{\link{influjew}},
\code{\link{influwht}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}}
}
\concept{Intergroup Relations}
\keyword{variable}
