% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{blnumoky}
\alias{blnumoky}
\title{Number of blacks in military-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
blnumoky
}
\details{
Question B. At the present time, about 33 percent of the army are black. All things considered, do you think there are too many blacks in the army, about the right number or should there be more blacks in the army?  RECORD UNDER "B."
}
\section{Values}{

\itemize{
\item \code{1} too many
\item \code{2} about right number
\item \code{3} should be more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/787/vshow}{GSS Data Explorer page for \code{blnumoky}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too many \tab about right number \tab should be more \tab don't know \tab no answer \tab Total \cr
   1982 \tab 923 \tab 115 \tab 555 \tab 153 \tab 106 \tab 8 \tab 1860 \cr
   Total \tab 923 \tab 115 \tab 555 \tab 153 \tab 106 \tab 8 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
