% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_148.R
\name{blues}
\alias{blues}
\title{Like or dislike blues or r and b music}
\source{
General Social Survey https://gss.norc.org
}
\description{
blues
}
\details{
Question I'm going to read you a list of some types of music. Can you tell me which of the statements on this card comes closest to your feeling about each type of music (HAND CARD CA TO RESPONDENT.) Let's start with big band music. Do you like it very much, like it, have mixed feelings, dislike it, dislike it very much, or is this a type of music that you don't know much about?
D. Blues or Rhythm and Blues
}
\section{Values}{

\itemize{
\item \verb{[1]} like very much
\item \verb{[2]} like it
\item \verb{[3]} mixed feelings
\item \verb{[4]} dislike it
\item \verb{[5]} dislike very much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1390/vshow}{GSS Data Explorer page for \code{blues}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab like very much \tab like it \tab mixed feelings \tab dislike it \tab dislike very much \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 221 \tab 669 \tab 367 \tab 220 \tab 61 \tab 56 \tab 12 \tab 1606 \cr
   Total \tab 0 \tab 221 \tab 669 \tab 367 \tab 220 \tab 61 \tab 56 \tab 12 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{classicl}},
\code{\link{country}},
\code{\link{folk}},
\code{\link{gospel}},
\code{\link{jazz}},
\code{\link{latin}},
\code{\link{moodeasy}},
\code{\link{musicals}},
\code{\link{newage}}
}
\concept{Culture}
\keyword{variable}
