% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{bookmark}
\alias{bookmark}
\title{How often r uses bookmarks}
\source{
General Social Survey https://gss.norc.org
}
\description{
bookmark
}
\details{
Question Now I’d like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit:
D. Use a bookmark or "favorites listing"?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than half of the time
\item \code{3} about half of the time
\item \code{4} more than half of the time
\item \code{5} almost always
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2391/vshow}{GSS Data Explorer page for \code{bookmark}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than half of the time \tab about half of the time \tab more than half of the time \tab almost always \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 276 \tab 140 \tab 92 \tab 80 \tab 68 \tab 86 \tab - \tab 2817 \cr
   2002 \tab 2403 \tab 62 \tab 89 \tab 40 \tab 56 \tab 65 \tab 49 \tab 1 \tab 2765 \cr
   Total \tab 4478 \tab 338 \tab 229 \tab 132 \tab 136 \tab 133 \tab 135 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{catdrcty}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{fin30}},
\code{\link{govt30}},
\code{\link{hyperlnk}},
\code{\link{news30}},
\code{\link{schl30}},
\code{\link{work30}}
}
\concept{Information Society}
\keyword{variable}
