% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{books16}
\alias{books16}
\title{Books in home of r age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
books16
}
\details{
Question About how many books were there around your family's house when you were 16 years old?
}
\section{Values}{

\itemize{
\item \verb{[1]} none
\item \verb{[2]} 1 or 2
\item \verb{[3]} around 10
\item \verb{[4]} around 20
\item \verb{[5]} around 50
\item \verb{[6]} around 100
\item \verb{[7]} around 200
\item \verb{[8]} around 500
\item \verb{[9]} 1,000 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4342/vshow}{GSS Data Explorer page for \code{books16}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab none \tab 1 or 2 \tab around 10 \tab around 20 \tab around 50 \tab around 100 \tab around 200 \tab around 500 \tab 1,000 or more \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2000 \tab 1545 \tab 37 \tab 81 \tab 205 \tab 226 \tab 229 \tab 186 \tab 139 \tab 63 \tab 33 \tab 1 \tab 72 \tab - \tab 2817 \cr
   Total \tab 1545 \tab 37 \tab 81 \tab 205 \tab 226 \tab 229 \tab 186 \tab 139 \tab 63 \tab 33 \tab 1 \tab 72 \tab 0 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{goodlife}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{inequal7}},
\code{\link{justpay}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{occmobil}},
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oppol}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}},
\code{\link{opsex}},
\code{\link{opwlth}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{pasup}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychild}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payexec}},
\code{\link{payfam1}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payhard}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payresp}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{paysup}},
\code{\link{payunskl}},
\code{\link{rank}()},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}},
\code{\link{unsklrnk}}

Other Social Class: 
\code{\link{bossemps}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{conage}},
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{educop}},
\code{\link{equal1}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{goveqinc}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{wlthpov}}
}
\concept{ISSP Social Inequality}
\concept{Social Class}
\keyword{variable}
