% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{bornsp}
\alias{bornsp}
\title{In what country were you born}
\source{
General Social Survey https://gss.norc.org
}
\description{
bornsp
}
\details{
Question In what country were you born?
}
\section{Values}{

\itemize{
\item \verb{[1]} usa
\item \verb{[2]} antigua
\item \verb{[3]} trinidad
\item \verb{[4]} jamaica
\item \verb{[5]} pakistan
\item \verb{[6]} turkey
\item \verb{[7]} italy
\item \verb{[8]} west indies
\item \verb{[9]} colombia
\item \verb{[10]} kazakhstan
\item \verb{[12]} cuba
\item \verb{[13]} india
\item \verb{[14]} germany
\item \verb{[15]} ecuador
\item \verb{[16]} mexico
\item \verb{[17]} portugal
\item \verb{[18]} el salvador
\item \verb{[19]} canada
\item \verb{[20]} guatemala
\item \verb{[21]} zimbabwe
\item \verb{[23]} taiwan
\item \verb{[24]} philippines
\item \verb{[25]} israel
\item \verb{[26]} dominican republic
\item \verb{[27]} scotland
\item \verb{[28]} haiti
\item \verb{[29]} honduras
\item \verb{[30]} puerto rico
\item \verb{[31]} austraila
\item \verb{[32]} china
\item \verb{[33]} poland
\item \verb{[34]} lithuania
\item \verb{[35]} liberia
\item \verb{[36]} brazil
\item \verb{[37]} morocco
\item \verb{[38]} spain
\item \verb{[39]} bulgeria
\item \verb{[40]} ussr
\item \verb{[41]} kuwait
\item \verb{[42]} japan
\item \verb{[43]} korea
\item \verb{[44]} vietnam
\item \verb{[45]} bermuda
\item \verb{[46]} ghana
\item \verb{[47]} panama
\item \verb{[48]} france
\item \verb{[49]} romania
\item \verb{[50]} south africa
\item \verb{[51]} england
\item \verb{[52]} britain
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3898/vshow}{GSS Data Explorer page for \code{bornsp}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab usa \tab antigua \tab trinidad \tab jamaica \tab pakistan \tab turkey \tab italy \tab west indies \tab colombia \tab kazakhstan \tab cuba \tab india \tab germany \tab ecuador \tab mexico \tab portugal \tab el salvador \tab canada \tab guatemala \tab zimbabwe \tab taiwan \tab philippines \tab israel \tab dominican republic \tab scotland \tab haiti \tab honduras \tab puerto rico \tab austraila \tab china \tab poland \tab lithuania \tab liberia \tab brazil \tab morocco \tab spain \tab bulgeria \tab ussr \tab kuwait \tab japan \tab korea \tab vietnam \tab bermuda \tab ghana \tab panama \tab france \tab romania \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 603 \tab 1179 \tab 1 \tab 5 \tab 3 \tab 3 \tab 2 \tab 3 \tab 2 \tab 6 \tab 2 \tab 8 \tab 10 \tab 9 \tab 1 \tab 60 \tab 1 \tab 9 \tab 6 \tab 6 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 56 \tab - \tab 2023 \cr
   Total \tab 603 \tab 1179 \tab 1 \tab 5 \tab 3 \tab 3 \tab 2 \tab 3 \tab 2 \tab 6 \tab 2 \tab 8 \tab 10 \tab 9 \tab 1 \tab 60 \tab 1 \tab 9 \tab 6 \tab 6 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 56 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{autonojb}},
\code{\link{curempyr}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{natdeg}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{typevisa}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work3yrs}}

Other Citizenship: 
\code{\link{actassoc}},
\code{\link{agecmeus}},
\code{\link{buypol}},
\code{\link{fucitzn}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obeylaws}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{typevisa}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{visa}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Nativity: 
\code{\link{born}},
\code{\link{granborn}},
\code{\link{migage}},
\code{\link{mighist}},
\code{\link{natborn}},
\code{\link{parborn}}
}
\concept{Citizenship}
\concept{Global Economics}
\concept{Immigration}
\concept{Nativity}
\keyword{variable}
