% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_455.R
\name{borrow2}
\alias{borrow2}
\title{Borrow money #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
borrow2
}
\details{
Question Suppose you needed to borrow a large sum of money.
B. Second Choice
}
\section{Values}{

\itemize{
\item \code{1} spouse
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other relative
\item \code{9} closest friend
\item \code{10} other friend
\item \code{11} neighbor
\item \code{12} coworker
\item \code{13} bank
\item \code{14} employer
\item \code{15} govt or soc ser
\item \code{16} other
\item \code{17} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4174/vshow}{GSS Data Explorer page for \code{borrow2}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab neighbor \tab coworker \tab bank \tab employer \tab govt or soc ser \tab other \tab no one \tab don't know \tab no answer \tab Total \cr
   1986 \tab 42 \tab 22 \tab 109 \tab 130 \tab 46 \tab 65 \tab 52 \tab 119 \tab 175 \tab 122 \tab 45 \tab 8 \tab 10 \tab 206 \tab 38 \tab 37 \tab 6 \tab 192 \tab 1 \tab 45 \tab 1470 \cr
   Total \tab 42 \tab 22 \tab 109 \tab 130 \tab 46 \tab 65 \tab 52 \tab 119 \tab 175 \tab 122 \tab 45 \tab 8 \tab 10 \tab 206 \tab 38 \tab 37 \tab 6 \tab 192 \tab 1 \tab 45 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{down1}},
\code{\link{down2}},
\code{\link{sick1}},
\code{\link{sick2}},
\code{\link{upset1}},
\code{\link{upset2}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
