% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{bosswrks}
\alias{bosswrks}
\title{Management & workers always in conflict}
\source{
General Social Survey https://gss.norc.org
}
\description{
bosswrks
}
\details{
Question How much do you agree or disagree with these two statements?
A. There will always be conflict between management and workers because they are really on opposite sides
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4505/vshow}{GSS Data Explorer page for \code{bosswrks}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 119 \tab 426 \tab 330 \tab 426 \tab 83 \tab 46 \tab 23 \tab 1537 \cr
   Total \tab 84 \tab 119 \tab 426 \tab 330 \tab 426 \tab 83 \tab 46 \tab 23 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{hiinc}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}},
\code{\link{strngun}}
}
\concept{Work Orientation}
\keyword{variable}
