% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{boyorgrl}
\alias{boyorgrl}
\title{Sci knowledge: father gene decides sex of baby}
\source{
General Social Survey https://gss.norc.org
}
\description{
boyorgrl
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don't know or aren't sure, just tell me so,  and we will skip to the next question. Remember true, false, or don't know.
C. It is the father's gene that decides whether the baby is a boy or a girl. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3336/vshow}{GSS Data Explorer page for \code{boyorgrl}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1175 \tab 371 \tab 317 \tab 1 \tab 4510 \cr
   2008 \tab 772 \tab 770 \tab 317 \tab 157 \tab 7 \tab 2023 \cr
   2010 \tab 1103 \tab 573 \tab 229 \tab 137 \tab 2 \tab 2044 \cr
   2012 \tab 972 \tab 586 \tab 260 \tab 149 \tab 7 \tab 1974 \cr
   2014 \tab 1299 \tab 713 \tab 336 \tab 188 \tab 2 \tab 2538 \cr
   2016 \tab 1477 \tab 833 \tab 368 \tab 180 \tab 9 \tab 2867 \cr
   2018 \tab 1173 \tab 679 \tab 341 \tab 148 \tab 7 \tab 2348 \cr
   Total \tab 9442 \tab 5329 \tab 2222 \tab 1276 \tab 35 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{electron}},
\code{\link{hotcore}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{radioact}},
\code{\link{viruses}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
