% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{brazil}
\alias{brazil}
\title{Liking for brazil}
\source{
General Social Survey https://gss.norc.org
}
\description{
brazil
}
\details{
Question You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? READ EACH ITEM.
E. Brazil
}
\section{Values}{

\itemize{
\item \code{0} +5
\item \code{1} +4
\item \code{2} +3
\item \code{3} +2
\item \code{4} +1
\item \code{5} -1
\item \code{6} -2
\item \code{7} -3
\item \code{8} -4
\item \code{9} -5
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974, 1975, 1977, 1982, 1983, 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5494/vshow}{GSS Data Explorer page for \code{brazil}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab +5 \tab +4 \tab +3 \tab +2 \tab +1 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab iap \tab Total \cr
   1974 \tab - \tab 97 \tab 126 \tab 256 \tab 196 \tab 333 \tab 135 \tab 47 \tab 36 \tab 21 \tab 24 \tab 213 \tab - \tab 1484 \cr
   1975 \tab - \tab 77 \tab 102 \tab 208 \tab 230 \tab 342 \tab 141 \tab 69 \tab 56 \tab 27 \tab 25 \tab 213 \tab - \tab 1490 \cr
   1977 \tab - \tab 92 \tab 94 \tab 194 \tab 195 \tab 338 \tab 161 \tab 76 \tab 65 \tab 39 \tab 68 \tab 208 \tab - \tab 1530 \cr
   1982 \tab - \tab 110 \tab 139 \tab 253 \tab 256 \tab 410 \tab 186 \tab 79 \tab 57 \tab 26 \tab 39 \tab 305 \tab - \tab 1860 \cr
   1983 \tab - \tab 74 \tab 72 \tab 153 \tab 250 \tab 475 \tab 205 \tab 78 \tab 55 \tab 22 \tab 29 \tab 186 \tab - \tab 1599 \cr
   1985 \tab - \tab 31 \tab 43 \tab 98 \tab 100 \tab 209 \tab 80 \tab 37 \tab 27 \tab 15 \tab 19 \tab 875 \tab - \tab 1534 \cr
   Total \tab 0 \tab 481 \tab 576 \tab 1162 \tab 1227 \tab 2107 \tab 908 \tab 386 \tab 296 \tab 150 \tab 204 \tab 2000 \tab 0 \tab 9497 \cr
}
}

\seealso{
Other War and International: 
\code{\link{canada}},
\code{\link{china}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{israel}},
\code{\link{japan}}
}
\concept{War and International}
\keyword{variable}
