% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{breakdwn}
\alias{breakdwn}
\title{How likely x is experiencing nervous breakdown}
\source{
General Social Survey https://gss.norc.org
}
\description{
breakdwn
}
\details{
Question In your opinion, how likely is it that NAME is experiencing (READ A-E) -- very likely, somewhat likely, not very likely, or not at all likely? (IN PART E, READ ONLY THE PHRASE FOLLOWING THE LETTER OF THE VIGNETTE)
B. A nervous breakdown
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1995/vshow}{GSS Data Explorer page for \code{breakdwn}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1460 \tab 249 \tab 501 \tab 411 \tab 207 \tab 52 \tab 24 \tab - \tab - \tab 2904 \cr
   2006 \tab 3073 \tab 269 \tab 530 \tab 387 \tab 185 \tab 51 \tab 15 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 206 \tab 420 \tab 328 \tab 162 \tab 28 \tab 29 \tab - \tab - \tab 2348 \cr
   2024 \tab 1691 \tab 225 \tab 655 \tab 481 \tab 193 \tab 31 \tab 12 \tab - \tab 21 \tab 3309 \cr
   Total \tab 7399 \tab 949 \tab 2106 \tab 1607 \tab 747 \tab 162 \tab 80 \tab 0 \tab 21 \tab 13071 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}

Other Mental Health: 
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
