% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{bribe}
\alias{bribe}
\title{Does r see public official asking for a bribe in last five years}
\source{
General Social Survey https://gss.norc.org
}
\description{
bribe
}
\details{
Question In the last five years, how often have you or a member of your immediate family come across a public official who hinted they wanted, or asked for, a bribe or favour in return for a service?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} seldom
\item \verb{[3]} occasionally
\item \verb{[4]} quite often
\item \verb{[5]} very often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4125/vshow}{GSS Data Explorer page for \code{bribe}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab seldom \tab occasionally \tab quite often \tab very often \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 1330 \tab 100 \tab 61 \tab 17 \tab 5 \tab 3 \tab 2 \tab 4510 \cr
   2016 \tab 1477 \tab 1174 \tab 105 \tab 65 \tab 17 \tab 12 \tab 4 \tab 13 \tab 2867 \cr
   Total \tab 4469 \tab 2504 \tab 205 \tab 126 \tab 34 \tab 17 \tab 7 \tab 15 \tab 7377 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{knowpols}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
