% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{brkdas1}
\alias{brkdas1}
\title{1st mention, what nervous breakdown is, asserted}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdas1
}
\details{
Question As far as you know, what is a nervous breakdown? (PROBES: How would you describe it?  What is it like?  What happens to a person who has one? How does he act?)
\enumerate{
\item R's asserted references for what a nervous breakdown is -- First mention.
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2049/vshow}{GSS Data Explorer page for \code{brkdas1}}.
}

\seealso{
Other Mental Health: 
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}

Other Mental Health: 
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
