% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{brkddo1}
\alias{brkddo1}
\title{1st mention what r did when felt ner. break}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkddo1
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
B. (IF EVER ANTICIPATED A NERVOUS BREAKDOWN) What did you do about it? (PROBE: Anything
else? until R says no.)  Which of these things did you do first; second; third?
(INTERVIEWER: Sequence list; read back if needed.)
B1. What R did about the nervous breakdown - First mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2070/vshow}{GSS Data Explorer page for \code{brkddo1}}.
}

\seealso{
Other Mental Health: 
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}}

Other Mental Health: 
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}}
}
\concept{Mental Health}
\keyword{variable}
