% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{brkdhlp1}
\alias{brkdhlp1}
\title{1st mention what help r sought}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdhlp1
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
C1. R''s sources of help during nervous breakdown -- First mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2073/vshow}{GSS Data Explorer page for \code{brkdhlp1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 100 \tab 101 \tab 102 \tab 103 \tab 105 \tab 106 \tab 107 \tab 108 \tab 190 \tab 200 \tab 201 \tab 300 \tab 390 \tab 400 \tab 401 \tab 402 \tab 403 \tab 408 \tab 480 \tab 490 \tab 499 \tab 500 \tab 502 \tab 503 \tab 505 \tab 506 \tab 604 \tab 605 \tab 606 \tab 610 \tab 690 \tab 700 \tab 799 \tab 800 \tab 860 \tab 910 \tab no answer \tab Total \cr
   1996 \tab 2517 \tab 12 \tab 2 \tab 1 \tab 8 \tab 3 \tab 23 \tab 4 \tab 1 \tab 3 \tab 20 \tab 13 \tab 1 \tab 1 \tab 4 \tab 4 \tab 2 \tab 38 \tab 1 \tab 2 \tab 2 \tab 3 \tab 10 \tab 6 \tab 1 \tab 1 \tab 38 \tab 3 \tab 8 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 131 \tab 29 \tab 2904 \cr
   Total \tab 2517 \tab 12 \tab 2 \tab 1 \tab 8 \tab 3 \tab 23 \tab 4 \tab 1 \tab 3 \tab 20 \tab 13 \tab 1 \tab 1 \tab 4 \tab 4 \tab 2 \tab 38 \tab 1 \tab 2 \tab 2 \tab 3 \tab 10 \tab 6 \tab 1 \tab 1 \tab 38 \tab 3 \tab 8 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 131 \tab 29 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdtime}},
\code{\link{evmhp}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
