% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{brkdhlp2}
\alias{brkdhlp2}
\title{2nd mention what help r sought}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdhlp2
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
C2. R''s sources of help during nervous breakdown -- Second mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2074/vshow}{GSS Data Explorer page for \code{brkdhlp2}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab 100 \tab 101 \tab 106 \tab 107 \tab 190 \tab 200 \tab 201 \tab 202 \tab 290 \tab 390 \tab 400 \tab 401 \tab 402 \tab 403 \tab 500 \tab 502 \tab 505 \tab 506 \tab 590 \tab 604 \tab 605 \tab 606 \tab 800 \tab 910 \tab no answer \tab Total \cr
   1996 \tab 2524 \tab 3 \tab 2 \tab 8 \tab 2 \tab 2 \tab 17 \tab 5 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 10 \tab 4 \tab 1 \tab 1 \tab 12 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab 260 \tab 30 \tab 2904 \cr
   Total \tab 2524 \tab 3 \tab 2 \tab 8 \tab 2 \tab 2 \tab 17 \tab 5 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 10 \tab 4 \tab 1 \tab 1 \tab 12 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab 260 \tab 30 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp3}},
\code{\link{brkdtime}},
\code{\link{evmhp}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
