% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{brkdhlp3}
\alias{brkdhlp3}
\title{3rd mention what help r sought}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdhlp3
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
C3. R''s sources of help during nervous breakdown -- Third mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2075/vshow}{GSS Data Explorer page for \code{brkdhlp3}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 100 \tab 103 \tab 106 \tab 108 \tab 200 \tab 300 \tab 390 \tab 400 \tab 401 \tab 500 \tab 506 \tab 606 \tab 710 \tab 800 \tab 910 \tab no answer \tab Total \cr
   1996 \tab 2533 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 319 \tab 33 \tab 2904 \cr
   Total \tab 2533 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 319 \tab 33 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdtime}},
\code{\link{evmhp}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
