% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{brkdimp2}
\alias{brkdimp2}
\title{2nd mention, what nervous breakdown is, imputed}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdimp2
}
\details{
Question As far as you know, what is a nervous breakdown? (PROBES: How would you describe it?  What is it like?  What happens to a person who has one? How does he act?)
2. R's imputed references for what a nervous breakdown is -- Second mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2052/vshow}{GSS Data Explorer page for \code{brkdimp2}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab 101 \tab 102 \tab 103 \tab 104 \tab 105 \tab 106 \tab 107 \tab 108 \tab 109 \tab 110 \tab 111 \tab 112 \tab 113 \tab 114 \tab no answer \tab Total \cr
   1996 \tab 2645 \tab 7 \tab 9 \tab 11 \tab 37 \tab 46 \tab 8 \tab 7 \tab 2 \tab 21 \tab 9 \tab 5 \tab 1 \tab 8 \tab 1 \tab 87 \tab 2904 \cr
   Total \tab 2645 \tab 7 \tab 9 \tab 11 \tab 37 \tab 46 \tab 8 \tab 7 \tab 2 \tab 21 \tab 9 \tab 5 \tab 1 \tab 8 \tab 1 \tab 87 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{knwmhosp}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}}
}
\concept{Mental Health}
\keyword{variable}
