% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{brkdsym3}
\alias{brkdsym3}
\title{3rd mention, what nervous breakdown is, symptoms}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdsym3
}
\details{
Question As far as you know, what is a nervous breakdown? (PROBES: How would you describe it?  What is it like?  What happens to a person who has one? How does he act?)
3. Symptoms or manifestations of nervous breakdown -- Third mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2056/vshow}{GSS Data Explorer page for \code{brkdsym3}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 201 \tab 202 \tab 203 \tab 204 \tab 205 \tab 207 \tab 210 \tab 211 \tab 212 \tab 302 \tab 303 \tab 304 \tab 305 \tab 306 \tab 307 \tab 309 \tab 310 \tab 311 \tab 405 \tab 407 \tab 411 \tab 501 \tab 502 \tab 503 \tab 506 \tab 508 \tab 511 \tab 512 \tab 601 \tab 602 \tab 609 \tab 610 \tab 611 \tab 612 \tab 701 \tab 702 \tab 704 \tab 706 \tab 707 \tab 709 \tab 710 \tab 803 \tab 807 \tab 808 \tab 811 \tab 812 \tab 902 \tab 904 \tab 905 \tab 906 \tab 907 \tab 908 \tab 911 \tab 921 \tab 923 \tab 925 \tab no answer \tab Total \cr
   1996 \tab 2531 \tab 1 \tab 10 \tab 2 \tab 1 \tab 12 \tab 14 \tab 1 \tab 2 \tab 5 \tab 1 \tab 7 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 11 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab 3 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab 15 \tab 5 \tab 1 \tab 3 \tab 4 \tab 11 \tab 21 \tab 6 \tab 12 \tab 1 \tab 27 \tab 8 \tab 14 \tab 9 \tab 3 \tab 6 \tab 1 \tab 96 \tab 2904 \cr
   Total \tab 2531 \tab 1 \tab 10 \tab 2 \tab 1 \tab 12 \tab 14 \tab 1 \tab 2 \tab 5 \tab 1 \tab 7 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 11 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab 3 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab 15 \tab 5 \tab 1 \tab 3 \tab 4 \tab 11 \tab 21 \tab 6 \tab 12 \tab 1 \tab 27 \tab 8 \tab 14 \tab 9 \tab 3 \tab 6 \tab 1 \tab 96 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{knwmhosp}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}}
}
\concept{Mental Health}
\keyword{variable}
