% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{brkdtime}
\alias{brkdtime}
\title{When feelings of nervous breakdown happened}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdtime
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
D. The timing of R''s nervous breakdown.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2076/vshow}{GSS Data Explorer page for \code{brkdtime}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab no answer \tab Total \cr
   1996 \tab 2517 \tab 17 \tab 77 \tab 96 \tab 98 \tab 99 \tab 2904 \cr
   Total \tab 2517 \tab 17 \tab 77 \tab 96 \tab 98 \tab 99 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{evmhp}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
