% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{brkdwhy1}
\alias{brkdwhy1}
\title{1st mention cause of ner. break feelings}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdwhy1
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
A1. (IF YES TO EVBRKDWN) Could you tell me about when you felt that way?  What was it about? (PROBE: GET EXACT DATE and complete details.)
The causes and sources of R''s nervous breakdown -- First mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2067/vshow}{GSS Data Explorer page for \code{brkdwhy1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 101 \tab 11 \tab 12 \tab 129 \tab 13 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 202 \tab 204 \tab 206 \tab 207 \tab 21 \tab 213 \tab 216 \tab 219 \tab 22 \tab 23 \tab 24 \tab 25 \tab 29 \tab 300 \tab 304 \tab 309 \tab 31 \tab 32 \tab 34 \tab 36 \tab 39 \tab 400 \tab 404 \tab 41 \tab 411 \tab 500 \tab 502 \tab 503 \tab 504 \tab 505 \tab 508 \tab 509 \tab 519 \tab 521 \tab 523 \tab 524 \tab 526 \tab 540 \tab 611 \tab 613 \tab 614 \tab 700 \tab 702 \tab 710 \tab 730 \tab 749 \tab 80 \tab 800 \tab 801 \tab 802 \tab 803 \tab 804 \tab 805 \tab 806 \tab 807 \tab 810 \tab 811 \tab 813 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2517 \tab 16 \tab 6 \tab 1 \tab 7 \tab 1 \tab 7 \tab 18 \tab 5 \tab 3 \tab 18 \tab 10 \tab 1 \tab 4 \tab 1 \tab 1 \tab 27 \tab 2 \tab 7 \tab 1 \tab 8 \tab 7 \tab 2 \tab 9 \tab 7 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 2 \tab 3 \tab 59 \tab 1 \tab 4 \tab 4 \tab 6 \tab 4 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 6 \tab 9 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab 4 \tab 1 \tab 11 \tab 7 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 30 \tab 2904 \cr
   Total \tab 2517 \tab 16 \tab 6 \tab 1 \tab 7 \tab 1 \tab 7 \tab 18 \tab 5 \tab 3 \tab 18 \tab 10 \tab 1 \tab 4 \tab 1 \tab 1 \tab 27 \tab 2 \tab 7 \tab 1 \tab 8 \tab 7 \tab 2 \tab 9 \tab 7 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 2 \tab 3 \tab 59 \tab 1 \tab 4 \tab 4 \tab 6 \tab 4 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 6 \tab 9 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab 4 \tab 1 \tab 11 \tab 7 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 30 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}}
}
\concept{Mental Health}
\keyword{variable}
