% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{brkdwhy2}
\alias{brkdwhy2}
\title{2nd mention cause of ner. break feelings}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdwhy2
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
A2. The causes and sources of R''s nervous breakdown -- Second mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2068/vshow}{GSS Data Explorer page for \code{brkdwhy2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 101 \tab 11 \tab 12 \tab 122 \tab 129 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 200 \tab 201 \tab 203 \tab 204 \tab 206 \tab 21 \tab 213 \tab 216 \tab 219 \tab 22 \tab 23 \tab 24 \tab 25 \tab 29 \tab 309 \tab 31 \tab 33 \tab 34 \tab 39 \tab 41 \tab 500 \tab 502 \tab 504 \tab 505 \tab 506 \tab 509 \tab 521 \tab 524 \tab 532 \tab 533 \tab 703 \tab 709 \tab 730 \tab 80 \tab 802 \tab 803 \tab 804 \tab 805 \tab 807 \tab 812 \tab 981 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2700 \tab 10 \tab 3 \tab 3 \tab 5 \tab 1 \tab 2 \tab 4 \tab 1 \tab 8 \tab 1 \tab 1 \tab 8 \tab 10 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 3 \tab 1 \tab 5 \tab 3 \tab 2 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 8 \tab 27 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 31 \tab 2904 \cr
   Total \tab 2700 \tab 10 \tab 3 \tab 3 \tab 5 \tab 1 \tab 2 \tab 4 \tab 1 \tab 8 \tab 1 \tab 1 \tab 8 \tab 10 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 3 \tab 1 \tab 5 \tab 3 \tab 2 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 8 \tab 27 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 31 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo1}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}}
}
\concept{Mental Health}
\keyword{variable}
