% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{brlawfl}
\alias{brlawfl}
\title{Do private companies comply with regulations}
\source{
General Social Survey https://gss.norc.org
}
\description{
brlawfl
}
\details{
Question In general, how often do you think that major private companies in America do the following? 
Comply with laws and regulations?  Would you say...
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6313/vshow}{GSS Data Explorer page for \code{brlawfl}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 290 \tab 392 \tab 513 \tab 149 \tab 32 \tab 14 \tab 2867 \cr
   Total \tab 1477 \tab 290 \tab 392 \tab 513 \tab 149 \tab 32 \tab 14 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
