% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{brneffrt}
\alias{brneffrt}
\title{Possible to become american if born elsewhere}
\source{
General Social Survey https://gss.norc.org
}
\description{
brneffrt
}
\details{
Question Some people say that it is possible to become truly American. if a person makes an effort. Others say a person has to be born American to be truly American. What is your position?
}
\section{Values}{

\itemize{
\item \verb{[1]} i definitely agree with statement a
\item \verb{[2]} i agree more with statement a than with statement b
\item \verb{[3]} i agree more with statement b than with statement a
\item \verb{[4]} i definitely agree with statement b
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8172/vshow}{GSS Data Explorer page for \code{brneffrt}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab i definitely agree with statement a \tab i agree more with statement a than with statement b \tab i agree more with statement b than with statement a \tab i definitely agree with statement b \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 933 \tab 350 \tab 130 \tab 68 \tab 144 \tab 1661 \tab 4 \tab 19 \tab 3309 \cr
   Total \tab 0 \tab 933 \tab 350 \tab 130 \tab 68 \tab 144 \tab 1661 \tab 4 \tab 19 \tab 3309 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immcrime1}},
\code{\link{immideas1}},
\code{\link{immjobs1}}
}
\concept{ISSP National Identity}
\keyword{variable}
