% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_451.R
\name{brovisit}
\alias{brovisit}
\title{How often visit brother}
\source{
General Social Survey https://gss.norc.org
}
\description{
brovisit
}
\details{
Question How many brothers age 18 or older do you have? (We mean brothers who are still alive; please include step-brothers, half-brothers, and adopted brothers.)
A. How often do you see or visit with your brother?
}
\section{Values}{

\itemize{
\item \code{1} lives in hshld
\item \code{2} daily
\item \code{3} several x week
\item \code{4} once a week
\item \code{5} once a month
\item \code{6} several x year
\item \code{7} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4140/vshow}{GSS Data Explorer page for \code{brovisit}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab lives in hshld \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab no answer \tab Total \cr
   1986 \tab 430 \tab 32 \tab 52 \tab 81 \tab 128 \tab 194 \tab 257 \tab 289 \tab 7 \tab 1470 \cr
   Total \tab 430 \tab 32 \tab 52 \tab 81 \tab 128 \tab 194 \tab 257 \tab 289 \tab 7 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{pacall}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
