% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_290.R
\name{browser2}
\alias{browser2}
\title{Can r give the name of a web browser}
\source{
General Social Survey https://gss.norc.org
}
\description{
browser2
}
\details{
Question C. Can you give me the name of a web browser?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2632/vshow}{GSS Data Explorer page for \code{browser2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2261 \tab 380 \tab 76 \tab 1 \tab 47 \tab 2765 \cr
   Total \tab 2261 \tab 380 \tab 76 \tab 1 \tab 47 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{askelse}},
\code{\link{askwksch}},
\code{\link{browser1}},
\code{\link{cussup}},
\code{\link{doself}},
\code{\link{manbook}},
\code{\link{payelse}},
\code{\link{processr}},
\code{\link{tecsup}}
}
\concept{Information Society}
\keyword{variable}
