% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_659.R
\name{brv16child}
\alias{brv16child}
\title{Death of child}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16child
}
\details{
Question We're very sorry. Who would that be?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7925/vshow}{GSS Data Explorer page for \code{brv16child}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no \tab yes \tab Total \cr
   2022 \tab 2767 \tab 241 \tab 536 \tab 3544 \cr
   Total \tab 2767 \tab 241 \tab 536 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16}},
\code{\link{brv16grand}},
\code{\link{brv16oth}},
\code{\link{brv16par}},
\code{\link{brv16sib}},
\code{\link{brv16sp}},
\code{\link{brv5cowknum}},
\code{\link{brv5frndnum}},
\code{\link{brv5othnum}}
}
\concept{GSS Next}
\keyword{variable}
