% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_661.R
\name{brv16chinum}
\alias{brv16chinum}
\title{Number of stepchild deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16chinum
}
\details{
Question NUMBER OF STEPCHILD DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7938/vshow}{GSS Data Explorer page for \code{brv16chinum}}.

Counts by year:\tabular{lll}{
   year \tab iap \tab Total \cr
   2022 \tab 3544 \tab 3544 \cr
   Total \tab 3544 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16frndnum}},
\code{\link{brv16othnum}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{gestate}}
}
\concept{GSS Next}
\keyword{variable}
