% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_660.R
\name{brv16filnum}
\alias{brv16filnum}
\title{Number of father in law deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16filnum
}
\details{
Question NUMBER OF FATHER IN LAW DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7932/vshow}{GSS Data Explorer page for \code{brv16filnum}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 0 \tab 1 \tab skipped on web \tab Total \cr
   2022 \tab 3440 \tab 54 \tab 6 \tab 44 \tab 3544 \cr
   Total \tab 3440 \tab 54 \tab 6 \tab 44 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16dadnum}},
\code{\link{brv16daunum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16partnum}},
\code{\link{brv16sonnum}},
\code{\link{brv16spnum}}
}
\concept{GSS Next}
\keyword{variable}
