% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_659.R
\name{brv16grand}
\alias{brv16grand}
\title{Death of grandparent}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16grand
}
\details{
Question We’re very sorry. Who would that be?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7924/vshow}{GSS Data Explorer page for \code{brv16grand}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no \tab yes \tab Total \cr
   2022 \tab 2767 \tab 733 \tab 44 \tab 3544 \cr
   Total \tab 2767 \tab 733 \tab 44 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16}},
\code{\link{brv16child}},
\code{\link{brv16oth}},
\code{\link{brv16par}},
\code{\link{brv16sib}},
\code{\link{brv16sp}},
\code{\link{brv5cowknum}},
\code{\link{brv5frndnum}},
\code{\link{brv5othnum}}
}
\concept{GSS Next}
\keyword{variable}
