% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_661.R
\name{brv16othnum}
\alias{brv16othnum}
\title{Number of other deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16othnum
}
\details{
Question NUMBER OF OTHER DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7944/vshow}{GSS Data Explorer page for \code{brv16othnum}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab skipped on web \tab Total \cr
   2022 \tab 3275 \tab 62 \tab 53 \tab 26 \tab 8 \tab 4 \tab 5 \tab 111 \tab 3544 \cr
   Total \tab 3275 \tab 62 \tab 53 \tab 26 \tab 8 \tab 4 \tab 5 \tab 111 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{brv16bronum}},
\code{\link{brv16chinum}},
\code{\link{brv16cuznum}},
\code{\link{brv16frndnum}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{gestate}}
}
\concept{GSS Next}
\keyword{variable}
