% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_660.R
\name{brv16partnum}
\alias{brv16partnum}
\title{Number of partner deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16partnum
}
\details{
Question NUMBER OF PARTNER DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[4]} one or more, number unknown
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7929/vshow}{GSS Data Explorer page for \code{brv16partnum}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 0 \tab 1 \tab 20 \tab skipped on web \tab Total \cr
   2022 \tab 3537 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3544 \cr
   Total \tab 3537 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16dadnum}},
\code{\link{brv16daunum}},
\code{\link{brv16filnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16sonnum}},
\code{\link{brv16spnum}}
}
\concept{GSS Next}
\keyword{variable}
