% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_660.R
\name{brv16sonnum}
\alias{brv16sonnum}
\title{Number of son deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16sonnum
}
\details{
Question NUMBER OF SON DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item No applicable value labels.
\item 
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7936/vshow}{GSS Data Explorer page for \code{brv16sonnum}}.

Counts by year:\tabular{lll}{
   year \tab iap \tab Total \cr
   2022 \tab 3544 \tab 3544 \cr
   Total \tab 3544 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16dadnum}},
\code{\link{brv16daunum}},
\code{\link{brv16filnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16partnum}},
\code{\link{brv16spnum}}
}
\concept{GSS Next}
\keyword{variable}
