% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_660.R
\name{brv16spnum}
\alias{brv16spnum}
\title{Number of spouse deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16spnum
}
\details{
Question NUMBER OF SPOUSE DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7928/vshow}{GSS Data Explorer page for \code{brv16spnum}}.
}

\seealso{
Other GSS Next: 
\code{\link{brv16dadnum}},
\code{\link{brv16filnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16grand}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16oth}},
\code{\link{brv16partnum}},
\code{\link{brv16sib}}
}
\concept{GSS Next}
\keyword{variable}
