% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_656.R
\name{brv5}
\alias{brv5}
\title{Death of someone close in past 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5
}
\details{
Question These next questions are about your experiences losing people close to you. Thinking back five years, to the period from January 1, 2017 through today, did you experience the death of someone close to you?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7896/vshow}{GSS Data Explorer page for \code{brv5}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2022 \tab 2309 \tab 874 \tab 361 \tab 3544 \cr
   Total \tab 2309 \tab 874 \tab 361 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{blkdsrv}},
\code{\link{blkmblty}},
\code{\link{blktry}},
\code{\link{brv5sp}}
}
\concept{GSS Next}
\keyword{variable}
