% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_659.R
\name{brv5frndnum}
\alias{brv5frndnum}
\title{Number of friend deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5frndnum
}
\details{
Question NUMBER OF FRIEND DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7918/vshow}{GSS Data Explorer page for \code{brv5frndnum}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 15 \tab 16 \tab 2 \tab 3 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab 9 \tab skipped on web \tab Total \cr
   2022 \tab 3070 \tab 40 \tab 132 \tab 1 \tab 1 \tab 2 \tab 1 \tab 75 \tab 50 \tab 18 \tab 24 \tab 1 \tab 8 \tab 2 \tab 3 \tab 1 \tab 115 \tab 3544 \cr
   Total \tab 3070 \tab 40 \tab 132 \tab 1 \tab 1 \tab 2 \tab 1 \tab 75 \tab 50 \tab 18 \tab 24 \tab 1 \tab 8 \tab 2 \tab 3 \tab 1 \tab 115 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16}},
\code{\link{brv16child}},
\code{\link{brv16grand}},
\code{\link{brv16oth}},
\code{\link{brv16par}},
\code{\link{brv16sib}},
\code{\link{brv16sp}},
\code{\link{brv5cowknum}},
\code{\link{brv5othnum}}
}
\concept{GSS Next}
\keyword{variable}
