% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_658.R
\name{brv5milnum}
\alias{brv5milnum}
\title{Number of mother in law deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5milnum
}
\details{
Question NUMBER OF MOTHER IN LAW DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7908/vshow}{GSS Data Explorer page for \code{brv5milnum}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab iap \tab skipped on web \tab Total \cr
   2022 \tab - \tab 121 \tab 45 \tab 2 \tab 3243 \tab 133 \tab 3544 \cr
   Total \tab 0 \tab 121 \tab 45 \tab 2 \tab 3243 \tab 133 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv5chinum}},
\code{\link{brv5dadnum}},
\code{\link{brv5daunum}},
\code{\link{brv5filnum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5momnum}},
\code{\link{brv5partnum}},
\code{\link{brv5sonnum}}
}
\concept{GSS Next}
\keyword{variable}
