% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_659.R
\name{brv5silnum}
\alias{brv5silnum}
\title{Number of brother-in-law or sister-in-law deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5silnum
}
\details{
Question NUMBER OF BROTHER-IN-LAW OR SISTER-IN-LAW DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7916/vshow}{GSS Data Explorer page for \code{brv5silnum}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab iap \tab skipped on web \tab Total \cr
   2022 \tab - \tab 30 \tab 31 \tab 8 \tab 2 \tab 1 \tab 3408 \tab 64 \tab 3544 \cr
   Total \tab 0 \tab 30 \tab 31 \tab 8 \tab 2 \tab 1 \tab 3408 \tab 64 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16}},
\code{\link{brv16par}},
\code{\link{brv16sp}},
\code{\link{brv5bronum}},
\code{\link{brv5cowknum}},
\code{\link{brv5cuznum}},
\code{\link{brv5frndnum}},
\code{\link{brv5othnum}},
\code{\link{brv5sisnum}}
}
\concept{GSS Next}
\keyword{variable}
