% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_603.R
\name{bthgrp2f}
\alias{bthgrp2f}
\title{What are the groups you both belong to?}
\source{
General Social Survey https://gss.norc.org
}
\description{
bthgrp2f
}
\details{
Question What are the groups you both belong to? Circle all that apply.
}
\section{Values}{

\itemize{
\item \code{1} fraternal groups
\item \code{2} service clubs
\item \code{3} veterans groups
\item \code{4} political clubs
\item \code{5} labor unions
\item \code{6} sports groups
\item \code{7} youth groups
\item \code{8} school service groups
\item \code{9} hobby or garden clubs
\item \code{10} school fraternities or sororities
\item \code{11} nationality groups
\item \code{12} farm organizations
\item \code{13} literary art discussion or study grps
\item \code{14} professional or academic societies
\item \code{15} church-affiliated groups
\item \code{16} other groups
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5526/vshow}{GSS Data Explorer page for \code{bthgrp2f}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab church-affiliated groups \tab no answer \tab Total \cr
   2004 \tab 2770 \tab 1 \tab 41 \tab 2812 \cr
   Total \tab 2770 \tab 1 \tab 41 \tab 2812 \cr
}
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
