% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_604.R
\name{bthgrp3f}
\alias{bthgrp3f}
\title{What are the groups you both belong to?}
\source{
General Social Survey https://gss.norc.org
}
\description{
bthgrp3f
}
\details{
Question What are the groups you both belong to? Circle all that apply.
}
\section{Values}{

\itemize{
\item \code{1} fraternal groups
\item \code{2} service clubs
\item \code{3} veterans groups
\item \code{4} political clubs
\item \code{5} labor unions
\item \code{6} sports groups
\item \code{7} youth groups
\item \code{8} school service groups
\item \code{9} hobby or garden clubs
\item \code{10} school fraternities or sororities
\item \code{11} nationality groups
\item \code{12} farm organizations
\item \code{13} literary art discussion or study grps
\item \code{14} professional or academic societies
\item \code{15} church-affiliated groups
\item \code{16} other groups
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5532/vshow}{GSS Data Explorer page for \code{bthgrp3f}}.
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp1a}},
\code{\link{bthgrp1b}},
\code{\link{bthgrp1c}},
\code{\link{bthgrp1d}},
\code{\link{bthgrp1e}},
\code{\link{bthgrp1f}},
\code{\link{bthgrp2a}},
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp2f}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}},
\code{\link{bthgrp4a}},
\code{\link{bthgrp4b}},
\code{\link{bthgrp4c}},
\code{\link{bthgrp4d}},
\code{\link{bthgrp4e}},
\code{\link{bthgrp4f}},
\code{\link{bthgrp5a}},
\code{\link{bthgrp5b}},
\code{\link{bthgrp5c}},
\code{\link{bthgrp5d}},
\code{\link{bthgrp5e}},
\code{\link{bthgrp5f}},
\code{\link{frstgrp1}},
\code{\link{frstgrp2}},
\code{\link{frstgrp3}},
\code{\link{frstgrp4}},
\code{\link{frstgrp5}},
\code{\link{inthisp}},
\code{\link{intrace1}},
\code{\link{intrace2}},
\code{\link{intrace3}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
