% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_604.R
\name{bthgrp4a}
\alias{bthgrp4a}
\title{What are the groups you both belong to?}
\source{
General Social Survey https://gss.norc.org
}
\description{
bthgrp4a
}
\details{
Question What are the groups you both belong to? Circle all that apply.
}
\section{Values}{

\itemize{
\item \code{1} fraternal groups
\item \code{2} service clubs
\item \code{3} veterans groups
\item \code{4} political clubs
\item \code{5} labor unions
\item \code{6} sports groups
\item \code{7} youth groups
\item \code{8} school service groups
\item \code{9} hobby or garden clubs
\item \code{10} school fraternities or sororities
\item \code{11} nationality groups
\item \code{12} farm organizations
\item \code{13} literary art discussion or study grps
\item \code{14} professional or academic societies
\item \code{15} church-affiliated groups
\item \code{16} other groups
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5533/vshow}{GSS Data Explorer page for \code{bthgrp4a}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab fraternal groups \tab sports groups \tab youth groups \tab school service groups \tab hobby or garden clubs \tab school fraternities or sororities \tab nationality groups \tab literary art discussion or study grps \tab professional or academic societies \tab church-affiliated groups \tab other groups \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2734 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 11 \tab 4 \tab 1 \tab 46 \tab 2812 \cr
   Total \tab 2734 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 11 \tab 4 \tab 1 \tab 46 \tab 2812 \cr
}
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp3f}},
\code{\link{bthgrp4b}},
\code{\link{bthgrp4c}},
\code{\link{bthgrp4d}},
\code{\link{bthgrp4e}},
\code{\link{bthgrp4f}},
\code{\link{bthgrp5a}},
\code{\link{bthgrp5b}},
\code{\link{bthgrp5c}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
