% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_572.R
\name{bugging}
\alias{bugging}
\title{People with secret should be subject to wiretapping}
\source{
General Social Survey https://gss.norc.org
}
\description{
bugging
}
\details{
Question Protecting secrets is a continuing concern of the government.
I am going to ask you about various security measures that the government might apply to individuals with a SECRET of TOP SECRET clearance. Do you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree that people with SECRET or TOP SECRET clearance should be subject to...
C. Wiretapping or electronic surveillance
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5245/vshow}{GSS Data Explorer page for \code{bugging}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 174 \tab 356 \tab 267 \tab 400 \tab 124 \tab 50 \tab 16 \tab 2832 \cr
   Total \tab 1445 \tab 174 \tab 356 \tab 267 \tab 400 \tab 124 \tab 50 \tab 16 \tab 2832 \cr
}
}

\seealso{
Other National Security: 
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{compfin}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{lietest}},
\code{\link{testdrug}}
}
\concept{National Security}
\keyword{variable}
