% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{bullied}
\alias{bullied}
\title{Situation caused by having been bullied}
\source{
General Social Survey https://gss.norc.org
}
\description{
bullied
}
\details{
Question In your opinion, how likely is it that (VIGNAME)’s situation might be caused by having been bullied?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8131/vshow}{GSS Data Explorer page for \code{bullied}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 345 \tab 795 \tab 321 \tab 119 \tab 36 \tab 1660 \tab 4 \tab 29 \tab 3309 \cr
   Total \tab 0 \tab 345 \tab 795 \tab 321 \tab 119 \tab 36 \tab 1660 \tab 4 \tab 29 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{imprvown1}},
\code{\link{morextrm}},
\code{\link{mustdoc1}},
\code{\link{musthosp1}},
\code{\link{mustmed1}},
\code{\link{smedia}}
}
\concept{Mental Health}
\keyword{variable}
