% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_634.R
\name{buseqinc}
\alias{buseqinc}
\title{Pay difference in companies}
\source{
General Social Survey https://gss.norc.org
}
\description{
buseqinc
}
\details{
Question Think of a score of 1 as meaning that major companies ought to reduce the pay differences between employees with high pay and those with low pay, and a score of 7 meaning that major companies should not concern themselves with reducing pay differences.
}
\section{Values}{

\itemize{
\item \code{1} should do something to reduce pay differences
\item \code{7} should not concern themselves with pay differences
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7546/vshow}{GSS Data Explorer page for \code{buseqinc}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab should do something to reduce pay differences \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab should not concern themselves with pay differences \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2241 \tab 547 \tab 196 \tab 316 \tab 372 \tab 156 \tab 70 \tab 97 \tab 5 \tab 4 \tab 28 \tab 4032 \cr
   Total \tab 2241 \tab 547 \tab 196 \tab 316 \tab 372 \tab 156 \tab 70 \tab 97 \tab 5 \tab 4 \tab 28 \tab 4032 \cr
}
}

\keyword{variable}
