% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{busing}
\alias{busing}
\title{Attitude toward racial busing}
\source{
General Social Survey https://gss.norc.org
}
\description{
busing
}
\details{
Question A. In general, do you favor or oppose the busing of (Negro/Black/African-American) and white school children from one school district to another?
}
\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/398/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab favor \tab no answer \tab oppose \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 62 \tab 313 \tab 7 \tab 1231 \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab 1504 \tab - \tab 1504 \cr
   1974 \tab 54 \tab 298 \tab 3 \tab 1129 \tab - \tab - \tab 1484 \cr
   1975 \tab 69 \tab 256 \tab 4 \tab 1161 \tab - \tab - \tab 1490 \cr
   1976 \tab 39 \tab 236 \tab 1 \tab 1223 \tab - \tab - \tab 1499 \cr
   1977 \tab 40 \tab 248 \tab 8 \tab 1234 \tab - \tab - \tab 1530 \cr
   1978 \tab 57 \tab 306 \tab 16 \tab 1153 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab 1468 \tab - \tab 1468 \cr
   1982 \tab 58 \tab 480 \tab 6 \tab 1316 \tab - \tab - \tab 1860 \cr
   1983 \tab 90 \tab 365 \tab 21 \tab 1123 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab 1473 \tab - \tab 1473 \cr
   1985 \tab 46 \tab 336 \tab 9 \tab 1143 \tab - \tab - \tab 1534 \cr
   1986 \tab 40 \tab 431 \tab 6 \tab 993 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab 1819 \tab - \tab 1819 \cr
   1988 \tab 44 \tab 320 \tab 3 \tab 621 \tab 493 \tab - \tab 1481 \cr
   1989 \tab 58 \tab 278 \tab 4 \tab 666 \tab 531 \tab - \tab 1537 \cr
   1990 \tab 80 \tab 451 \tab 2 \tab 839 \tab - \tab - \tab 1372 \cr
   1991 \tab 55 \tab 350 \tab 8 \tab 611 \tab 493 \tab - \tab 1517 \cr
   1993 \tab 86 \tab 310 \tab 4 \tab 680 \tab 526 \tab - \tab 1606 \cr
   1994 \tab 107 \tab 623 \tab 11 \tab 1236 \tab 1015 \tab - \tab 2992 \cr
   1996 \tab 62 \tab 350 \tab 3 \tab 576 \tab 1913 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 1047 \tab 5951 \tab 116 \tab 16935 \tab 44797 \tab 6853 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1988 \tab A/B/- \tab partial \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
}
}

\seealso{
Other Core: 
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}

Other Race: 
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
