% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{busing10}
\alias{busing10}
\title{Favor busing 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
busing10
}
\details{
Question B. Now, thinking about ten years ago, that is in 1972, did you then favor or oppose the busing of (Negro/Black) and white school children from one school district to another?
}
\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/399/vshow}{GSS Data Explorer page for \code{busing10}}.
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\keyword{variable}
