% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{buspow}
\alias{buspow}
\title{Power of business and industry}
\source{
General Social Survey https://gss.norc.org
}
\description{
buspow
}
\details{
Question How about business and industry, do they have too much power or too little power?
}
\section{Values}{

\itemize{
\item \code{1} far too much power
\item \code{2} too much power
\item \code{3} about the right amount of power
\item \code{4} too little power
\item \code{5} far too little power
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1996 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4069/vshow}{GSS Data Explorer page for \code{buspow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab far too much power \tab too much power \tab about the right amount of power \tab too little power \tab far too little power \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 79 \tab 221 \tab 269 \tab 36 \tab 3 \tab 61 \tab 8 \tab 1534 \cr
   1990 \tab 155 \tab 105 \tab 388 \tab 509 \tab 68 \tab 2 \tab 121 \tab 24 \tab 1372 \cr
   1996 \tab 1572 \tab 146 \tab 434 \tab 507 \tab 79 \tab 5 \tab 112 \tab 49 \tab 2904 \cr
   Total \tab 2584 \tab 330 \tab 1043 \tab 1285 \tab 183 \tab 10 \tab 294 \tab 81 \tab 5810 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
