% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_442.R
\name{bustax}
\alias{bustax}
\title{Amount of tax on business and industry}
\source{
General Social Survey https://gss.norc.org
}
\description{
bustax
}
\details{
Question Do you consider the amount of tax that business and industry have to pay is too high or too low?
}
\section{Values}{

\itemize{
\item \verb{[1]} much too high
\item \verb{[2]} too high
\item \verb{[3]} about right
\item \verb{[4]} too low
\item \verb{[5]} much too low
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4066/vshow}{GSS Data Explorer page for \code{bustax}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much too high \tab too high \tab about right \tab too low \tab much too low \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 32 \tab 82 \tab 199 \tab 217 \tab 67 \tab 71 \tab 9 \tab 1534 \cr
   Total \tab 857 \tab 32 \tab 82 \tab 199 \tab 217 \tab 67 \tab 71 \tab 9 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
