% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_197.R
\name{buyfixes}
\alias{buyfixes}
\title{Would you buy home maintenance from a business}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyfixes
}
\details{
Question I'd like you to imagine that you want to buy each of the following items or services. Would you rather buy them from a business or service provider with whom you had some family or social relationship, or from a seller or provider with whom you had no prior personal contact. Please tell me the number that best represents your feeling with respect to each of the purchases listed.
D. Home maintenance (e.g., carpentry, painting)
}
\section{Values}{

\itemize{
\item \code{1} strong preference no contact
\item \code{3} dont care
\item \code{5} strong preference contact
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1860/vshow}{GSS Data Explorer page for \code{buyfixes}}.
}

\seealso{
Other Markets: 
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{sellauto}},
\code{\link{sellhome}},
\code{\link{transoth}}
}
\concept{Markets}
\keyword{variable}
