% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_278.R
\name{buyinfgn}
\alias{buyinfgn}
\title{R sought information purchase past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyinfgn
}
\details{
Question In the last twelve months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/2002–have you:
A. Looked for information about a product you might want to buy?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2532/vshow}{GSS Data Explorer page for \code{buyinfgn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 1884 \tab 360 \tab 302 \tab 253 \tab 18 \tab - \tab - \tab 2817 \cr
   2002 \tab 1372 \tab 588 \tab 215 \tab 234 \tab 5 \tab 341 \tab 10 \tab 2765 \cr
   Total \tab 3256 \tab 948 \tab 517 \tab 487 \tab 23 \tab 341 \tab 10 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{gamegn}},
\code{\link{investgn}},
\code{\link{newjobgn}},
\code{\link{peoplegn}},
\code{\link{polcamp}},
\code{\link{poldisgn}},
\code{\link{polfrel}},
\code{\link{polwww}},
\code{\link{travelgn}}
}
\concept{Information Society}
\keyword{variable}
