% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_197.R
\name{buylaw}
\alias{buylaw}
\title{Would you buy legal svcs from a business}
\source{
General Social Survey https://gss.norc.org
}
\description{
buylaw
}
\details{
Question I'd like you to imagine that you want to buy each of the following items or services. Would you rather buy them from a business or service provider with whom you had some family or social relationship, or from a seller or provider with whom you had no prior personal contact. Please tell me the number that best represents your feeling with respect to each of the purchases listed.
C. Legal services
}
\section{Values}{

\itemize{
\item \code{1} strong preference no contact
\item \code{3} dont care
\item \code{5} strong preference contact
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1859/vshow}{GSS Data Explorer page for \code{buylaw}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strong preference no contact \tab 2 \tab dont care \tab 4 \tab strong preference contact \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2196 \tab 87 \tab 25 \tab 159 \tab 136 \tab 286 \tab 12 \tab 3 \tab 2904 \cr
   Total \tab 2196 \tab 87 \tab 25 \tab 159 \tab 136 \tab 286 \tab 12 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{sellauto}},
\code{\link{sellhome}},
\code{\link{transoth}}
}
\concept{Markets}
\keyword{variable}
