% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{byemprob}
\alias{byemprob}
\title{How many friends discuss problems via email}
\source{
General Social Survey https://gss.norc.org
}
\description{
byemprob
}
\details{
Question A. Of these (NUMPROBS) friends and relatives that you feel really close to, how many do you stay in contact with  through electronic mail or instant messaging?
}
\section{Values}{

\itemize{
\item \code{96} 96 or higher
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2607/vshow}{GSS Data Explorer page for \code{byemprob}}.
}

\seealso{
Other Information Society: 
\code{\link{intaccss}},
\code{\link{jobinfo}},
\code{\link{jobpaper}},
\code{\link{numprobs}}

Other Family: 
\code{\link{numprobs}}
}
\concept{Family}
\concept{Information Society}
\keyword{variable}
