% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{camwrk}
\alias{camwrk}
\title{Govt camera surveillance at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
camwrk
}
\details{
Question Protecting secrets is a continuing concern of the government. I am going to ask you about various security measures that the government might apply to individuals with a SECRET or TOP SECRET clearance. Do you (strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree) that people with SECRET or TOP SECRET clearance should be subject to:
H. Video camera surveillance in the workplace
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5256/vshow}{GSS Data Explorer page for \code{camwrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 274 \tab 627 \tab 219 \tab 182 \tab 70 \tab 30 \tab 17 \tab 2817 \cr
   Total \tab 1398 \tab 274 \tab 627 \tab 219 \tab 182 \tab 70 \tab 30 \tab 17 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{nuclrwar}},
\code{\link{srchwrk}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{usspy}},
\code{\link{usterror}}
}
\concept{National Security}
\keyword{variable}
