% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{caninf11}
\alias{caninf11}
\title{Would go to govt site for information on candid}
\source{
General Social Survey https://gss.norc.org
}
\description{
caninf11
}
\details{
Question Imagine you wanted to use the Web to find out information about a candidate for political office. What would you do?
K. Would go to government site for information on candidate.
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[2]} yes, at least one url is not a .com
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2994/vshow}{GSS Data Explorer page for \code{caninf11}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1789 \tab 686 \tab 9 \tab 25 \tab 308 \tab 2817 \cr
   Total \tab 1789 \tab 686 \tab 9 \tab 25 \tab 308 \tab 2817 \cr
}
}

\seealso{
Other Information Seeking Online: 
\code{\link{caninf12}},
\code{\link{caninf13}},
\code{\link{caninf14}}
}
\concept{Information Seeking Online}
\keyword{variable}
