% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{caninf12}
\alias{caninf12}
\title{Would refuse to answer not sure don't know}
\source{
General Social Survey https://gss.norc.org
}
\description{
caninf12
}
\details{
Question Imagine you wanted to use the Web to find out information about a candidate for political office. What would you do?
L. Would refuse to answer, not sure, or don''t know.
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{2} yes, at least one url is not a .com
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2995/vshow}{GSS Data Explorer page for \code{caninf12}}.
}

\seealso{
Other Information Seeking Online: 
\code{\link{caninf11}},
\code{\link{caninf13}},
\code{\link{caninf14}}

Other Political: 
\code{\link{caninf11}},
\code{\link{caninf13}},
\code{\link{caninf14}}
}
\concept{Information Seeking Online}
\concept{Political}
\keyword{variable}
