% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{caninf14}
\alias{caninf14}
\title{First site mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
caninf14
}
\details{
Question Imagine you wanted to use the Web to find out information about a candidate for political office. What would you do?
N. First site mentioned.
}
\section{Values}{

\itemize{
\item \verb{[0]} no specific search engine mentioned
\item \verb{[1]} yahoo
\item \verb{[2]} aol
\item \verb{[3]} cnn
\item \verb{[4]} msn
\item \verb{[5]} excite
\item \verb{[6]} google
\item \verb{[7]} webferret
\item \verb{[8]} lycos
\item \verb{[9]} altavista
\item \verb{[10]} usatoday
\item \verb{[11]} copernic
\item \verb{[12]} askjeeves
\item \verb{[13]} peoplefind.com
\item \verb{[14]} abc
\item \verb{[15]} msnbc
\item \verb{[16]} infoseek go
\item \verb{[17]} dogpile
\item \verb{[18]} snap
\item \verb{[19]} metacrawler
\item \verb{[20]} webcrawler
\item \verb{[21]} netscape
\item \verb{[22]} iwon
\item \verb{[23]} politics.com
\item \verb{[24]} homp page search engine
\item \verb{[25]} nbc
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2997/vshow}{GSS Data Explorer page for \code{caninf14}}.

Counts by year:\tabular{llllllllllllllllllllllllllllll}{
   year \tab iap \tab no specific search engine mentioned \tab yahoo \tab aol \tab cnn \tab msn \tab excite \tab google \tab webferret \tab lycos \tab altavista \tab usatoday \tab copernic \tab peoplefind.com \tab abc \tab msnbc \tab infoseek go \tab dogpile \tab snap \tab metacrawler \tab webcrawler \tab netscape \tab iwon \tab politics.com \tab homp page search engine \tab nbc \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1789 \tab 527 \tab 83 \tab 18 \tab 13 \tab 5 \tab 5 \tab 7 \tab 1 \tab 3 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 25 \tab 308 \tab 2817 \cr
   Total \tab 1789 \tab 527 \tab 83 \tab 18 \tab 13 \tab 5 \tab 5 \tab 7 \tab 1 \tab 3 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 25 \tab 308 \tab 2817 \cr
}
}

\seealso{
Other Information Seeking Online: 
\code{\link{caninf11}},
\code{\link{caninf12}},
\code{\link{caninf13}}
}
\concept{Information Seeking Online}
\keyword{variable}
