% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_334.R
\name{caninf9}
\alias{caninf9}
\title{Would go to news site for information on candid}
\source{
General Social Survey https://gss.norc.org
}
\description{
caninf9
}
\details{
Question Imagine you wanted to use the Web to find out information about a candidate for political office. What would you do?
I. Would go to news or media site for information on candidate.
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{2} yes, at least one url is not a .com
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2992/vshow}{GSS Data Explorer page for \code{caninf9}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1789 \tab 646 \tab 49 \tab 25 \tab 308 \tab 2817 \cr
   Total \tab 1789 \tab 646 \tab 49 \tab 25 \tab 308 \tab 2817 \cr
}
}

\seealso{
Other Information Seeking Online: 
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf6}},
\code{\link{caninf7}},
\code{\link{caninf8}}
}
\concept{Information Seeking Online}
\keyword{variable}
