% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{cantrust}
\alias{cantrust}
\title{People can be trusted or cant be too careful}
\source{
General Social Survey https://gss.norc.org
}
\description{
cantrust
}
\details{
Question Generally speaking, would you say that people can be trusted or that you can't be too careful in dealing with people?
}
\section{Values}{

\itemize{
\item \verb{[1]} people can almost always be trusted
\item \verb{[2]} people can usually be trusted
\item \verb{[3]} you usually can't be too careful in dealing with people
\item \verb{[4]} you almost always can't be too careful in dealing with people
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2004, 2014 \cr
   B/C/- \tab 2008, 2018, 2021, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4686/vshow}{GSS Data Explorer page for \code{cantrust}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab people can almost always be trusted \tab people can usually be trusted \tab you usually can't be too careful in dealing with people \tab you almost always can't be too careful in dealing with people \tab can't choose \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   1998 \tab 1548 \tab 67 \tab 538 \tab 494 \tab 117 \tab 37 \tab 31 \tab - \tab - \tab 2832 \cr
   2004 \tab 1340 \tab 55 \tab 650 \tab 617 \tab 142 \tab 1 \tab 7 \tab - \tab - \tab 2812 \cr
   2008 \tab 658 \tab 146 \tab 457 \tab 587 \tab 170 \tab 4 \tab 1 \tab - \tab - \tab 2023 \cr
   2014 \tab 1274 \tab 37 \tab 427 \tab 628 \tab 146 \tab 22 \tab 4 \tab - \tab - \tab 2538 \cr
   2018 \tab 1175 \tab 64 \tab 411 \tab 549 \tab 136 \tab 10 \tab 3 \tab - \tab - \tab 2348 \cr
   2021 \tab 2029 \tab 71 \tab 714 \tab 750 \tab 205 \tab 77 \tab 177 \tab 9 \tab - \tab 4032 \cr
   2024 \tab 1669 \tab 53 \tab 515 \tab 747 \tab 201 \tab 95 \tab 4 \tab 25 \tab - \tab 3309 \cr
   Total \tab 9693 \tab 493 \tab 3712 \tab 4372 \tab 1117 \tab 246 \tab 227 \tab 34 \tab 0 \tab 19894 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{befair}},
\code{\link{bible1}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conbiz}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{hapunhap}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{madenkid}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relcand}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}

Other Anomia: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{selfirst}}

Other People: 
\code{\link{advantge}},
\code{\link{befair}},
\code{\link{fair}},
\code{\link{fair5}},
\code{\link{helpful}},
\code{\link{trust}},
\code{\link{trust5}},
\code{\link{trusting}},
\code{\link{trustpeo}},
\code{\link{trusty}},
\code{\link{twoclass}},
\code{\link{wantbest}}
}
\concept{Anomia}
\concept{ISSP Religion}
\concept{People}
\keyword{variable}
