% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{capimp}
\alias{capimp}
\title{Importance of death penalty issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
capimp
}
\details{
Question How important is the death penalty issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \code{1} one of most imp
\item \code{2} important
\item \code{3} not very imp
\item \code{4} not important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/269/vshow}{GSS Data Explorer page for \code{capimp}}.
}

\seealso{
Other Core: 
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}
}
\concept{Core}
\keyword{variable}
