% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{capinfo}
\alias{capinfo}
\title{How much info r has on death penalty}
\source{
General Social Survey https://gss.norc.org
}
\description{
capinfo
}
\details{
Question How much information do you have about the death penalty issue? Do you have all of the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \code{1} all info r needs
\item \code{2} most of the info
\item \code{3} some info
\item \code{4} very little info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/270/vshow}{GSS Data Explorer page for \code{capinfo}}.
}

\seealso{
Other Core: 
\code{\link{capimp}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}

Other Capital Punishment: 
\code{\link{capimp}},
\code{\link{cappun}},
\code{\link{cappun2}}
}
\concept{Capital Punishment}
\concept{Core}
\keyword{variable}
