% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{cappun2}
\alias{cappun2}
\title{Does r favor capital punishment}
\source{
General Social Survey https://gss.norc.org
}
\description{
cappun2
}
\details{
Question Are you in favor of the death penalty for persons convicted of murder?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes,favor
\item \verb{[2]} no,dont favor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/267/vshow}{GSS Data Explorer page for \code{cappun2}}.

Counts by year:\tabular{lllllll}{
   year \tab yes,favor \tab no,dont favor \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 852 \tab 632 \tab 124 \tab 1 \tab 4 \tab 1613 \cr
   1973 \tab 898 \tab 519 \tab 75 \tab - \tab 12 \tab 1504 \cr
   Total \tab 1750 \tab 1151 \tab 199 \tab 1 \tab 16 \tab 3117 \cr
}
}

\seealso{
Other Core: 
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}
}
\concept{Core}
\keyword{variable}
