% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_192.R
\name{carbuy1}
\alias{carbuy1}
\title{What was the most important reason?}
\source{
General Social Survey https://gss.norc.org
}
\description{
carbuy1
}
\details{
Question What made you think that this would be a good place to make your purchase? (Choose all that apply.)
H. MOST IMPORTANT REASON FROM ABOVE
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1809/vshow}{GSS Data Explorer page for \code{carbuy1}}.
}

\seealso{
Other Markets: 
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{reldealr}},
\code{\link{satcar}}
}
\concept{Markets}
\keyword{variable}
