% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_192.R
\name{carbuyb}
\alias{carbuyb}
\title{A friend or relative recommended r buy}
\source{
General Social Survey https://gss.norc.org
}
\description{
carbuyb
}
\details{
Question What made you think that this would be a good place to make your purchase? (Choose all that apply.)
B. A friend or relative recommended them
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1803/vshow}{GSS Data Explorer page for \code{carbuyb}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2497 \tab 111 \tab 204 \tab 1 \tab 91 \tab 2904 \cr
   Total \tab 2497 \tab 111 \tab 204 \tab 1 \tab 91 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{reldealr}},
\code{\link{satcar}}
}
\concept{Markets}
\keyword{variable}
