% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_192.R
\name{carbuyc}
\alias{carbuyc}
\title{An acquaintance recommended r buy}
\source{
General Social Survey https://gss.norc.org
}
\description{
carbuyc
}
\details{
Question What made you think that this would be a good place to make your purchase? (Choose all that apply.)
C. An acquaintance recommended them
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1804/vshow}{GSS Data Explorer page for \code{carbuyc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2497 \tab 38 \tab 267 \tab 1 \tab 101 \tab 2904 \cr
   Total \tab 2497 \tab 38 \tab 267 \tab 1 \tab 101 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{reldealr}},
\code{\link{satcar}}
}
\concept{Markets}
\keyword{variable}
