% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{cardealr}
\alias{cardealr}
\title{Purchased a car from a dealership past 5 yrs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
cardealr
}
\details{
Question Have you purchased a car from a dealership in the past five years?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1799/vshow}{GSS Data Explorer page for \code{cardealr}}.
}

\seealso{
Other Markets: 
\code{\link{carprivt}},
\code{\link{newused}},
\code{\link{relprivt}},
\code{\link{typdealr}},
\code{\link{warrntyc}}

Other Purchases/Transactions: 
\code{\link{carprivt}},
\code{\link{newused}},
\code{\link{relprivt}},
\code{\link{typdealr}},
\code{\link{warrntyc}}
}
\concept{Markets}
\concept{Purchases/Transactions}
\keyword{variable}
