% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{cardo}
\alias{cardo}
\title{What would you do in this situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
cardo
}
\details{
Question Suppose you were riding in a car driven by a close friend. You know he is going too fast. He hits a pedestrian. He asks you to tell the police that he was obeying the speed limit.
B. What would you do in this situation?
}
\section{Values}{

\itemize{
\item \code{1} def tll plce was
\item \code{2} prb tll plce was
\item \code{3} prb tll plce not
\item \code{4} def tll plce not
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4697/vshow}{GSS Data Explorer page for \code{cardo}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{carright}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{ISSP Religion}
\keyword{variable}
