% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{caregrp}
\alias{caregrp}
\title{How committed to group is r}
\source{
General Social Survey https://gss.norc.org
}
\description{
caregrp
}
\details{
Question Personally, how committed do you feel to this organization?  Not committed at all, slightly committed, moderately committed, strongly committed, or very strongly committed.
}
\section{Values}{

\itemize{
\item \code{1} not committed
\item \code{2} slightly comm
\item \code{3} moderately comm
\item \code{4} strongly comm
\item \code{5} very strongly comm
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1014/vshow}{GSS Data Explorer page for \code{caregrp}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{churhgrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{mostact}},
\code{\link{servegrp}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
