% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{caremost}
\alias{caremost}
\title{Most concerned possible effect of global warming}
\source{
General Social Survey https://gss.norc.org
}
\description{
caremost
}
\details{
Question Please look at Card B24. Scientists predict that global warming may soon have big effects on the polar regions. I will describe some of these possible effects and, for each one, please say whether it would bother you a great deal, some, a little, or not at all if it actually happened.
G. Which of these possible effects do you care about most?
}
\section{Values}{

\itemize{
\item \code{1} the extinction of the polar bears
\item \code{2} the rise in sea level
\item \code{3} the threat to the arctic seals
\item \code{4} the threat to the inuit way of life
\item \code{5} the melting of the northern ice cap
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3444/vshow}{GSS Data Explorer page for \code{caremost}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab the extinction of the polar bears \tab the rise in sea level \tab the threat to the arctic seals \tab the threat to the inuit way of life \tab the melting of the northern ice cap \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 155 \tab 762 \tab 36 \tab 244 \tab 594 \tab 66 \tab 7 \tab 4510 \cr
   2010 \tab 1338 \tab 74 \tab 308 \tab 8 \tab 94 \tab 182 \tab 26 \tab 14 \tab 2044 \cr
   Total \tab 3984 \tab 229 \tab 1070 \tab 44 \tab 338 \tab 776 \tab 92 \tab 21 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
