% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{careself}
\alias{careself}
\title{Those in need have to take care of themselves}
\source{
General Social Survey https://gss.norc.org
}
\description{
careself
}
\details{
Question Please tell me whether you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree with the following statements:
B. Those in need have to learn to take care of themselves and not depend on others
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2891/vshow}{GSS Data Explorer page for \code{careself}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 175 \tab 552 \tab 319 \tab 254 \tab 52 \tab 13 \tab 7 \tab 2765 \cr
   2004 \tab 1472 \tab 115 \tab 519 \tab 393 \tab 255 \tab 50 \tab 6 \tab 2 \tab 2812 \cr
   2012 \tab 666 \tab 154 \tab 521 \tab 333 \tab 256 \tab 27 \tab 10 \tab 7 \tab 1974 \cr
   2014 \tab 1264 \tab 156 \tab 519 \tab 303 \tab 240 \tab 46 \tab 3 \tab 7 \tab 2538 \cr
   Total \tab 4795 \tab 600 \tab 2111 \tab 1348 \tab 1005 \tab 175 \tab 32 \tab 23 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selffrst}},
\code{\link{selfless}}
}
\concept{Altruism}
\keyword{variable}
