% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{carprivt}
\alias{carprivt}
\title{Have you purchased a used car past 5 yrs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
carprivt
}
\details{
Question Have you purchased a used car from a private individual in the past five years?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1796/vshow}{GSS Data Explorer page for \code{carprivt}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2168 \tab 250 \tab 485 \tab 1 \tab 2904 \cr
   Total \tab 2168 \tab 250 \tab 485 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{cardealr}},
\code{\link{newused}},
\code{\link{relprivt}},
\code{\link{typdealr}},
\code{\link{warrntyc}}
}
\concept{Markets}
\keyword{variable}
