% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{carright}
\alias{carright}
\title{What your friend has a right to expect from you}
\source{
General Social Survey https://gss.norc.org
}
\description{
carright
}
\details{
Question Suppose you were riding in a car driven by a close friend. You know he is going too fast. He hits a pedestrian. He asks you to tell the police that he was obeying the speed limit.
A. Which statement comes closest to your belief about what your friend has a right to expect from you?
}
\section{Values}{

\itemize{
\item \code{1} definite right
\item \code{2} some right
\item \code{3} no right
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4696/vshow}{GSS Data Explorer page for \code{carright}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab definite right \tab some right \tab no right \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 23 \tab 115 \tab 972 \tab 119 \tab 55 \tab 2832 \cr
   Total \tab 1548 \tab 23 \tab 115 \tab 972 \tab 119 \tab 55 \tab 2832 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{ISSP Religion}
\keyword{variable}
