% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{carsgen}
\alias{carsgen}
\title{Car pollution danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
carsgen
}
\details{
Question Please check one box for each of the questions below to show what you think.
A. In general, do you think that air pollution caused by cars is…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4753/vshow}{GSS Data Explorer page for \code{carsgen}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 265 \tab 472 \tab 584 \tab 106 \tab 10 \tab 34 \tab 86 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 222 \tab 363 \tab 581 \tab 102 \tab 10 \tab 37 \tab 71 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 232 \tab 347 \tab 521 \tab 68 \tab 6 \tab 48 \tab 54 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 228 \tab 411 \tab 627 \tab 105 \tab 13 \tab 38 \tab 8 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 232 \tab 559 \tab 814 \tab 151 \tab 22 \tab 59 \tab 184 \tab 8 \tab 4032 \cr
   Total \tab 5813 \tab 1179 \tab 2152 \tab 3127 \tab 532 \tab 61 \tab 216 \tab 403 \tab 8 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
