% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{carsten}
\alias{carsten}
\title{Car pollution will increase in next 10 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
carsten
}
\details{
Question Please check one box for each of the questions below to show what you think.
C. Within the next ten years, how likely do you think it is that there will be a large increase in ill-health In America's cities as a result of air pollution caused by cars?
}
\section{Values}{

\itemize{
\item \verb{[1]} certain to happen
\item \verb{[2]} very likely to happen
\item \verb{[3]} fairly likely to happen
\item \verb{[4]} not very likely to happen
\item \verb{[5]} not at all likely
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4755/vshow}{GSS Data Explorer page for \code{carsten}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab certain to happen \tab very likely to happen \tab fairly likely to happen \tab not very likely to happen \tab not at all likely \tab can't choose \tab no answer \tab Total \cr
   1993 \tab 49 \tab 214 \tab 429 \tab 441 \tab 302 \tab 11 \tab 73 \tab 87 \tab 1606 \cr
   1994 \tab 1606 \tab 150 \tab 401 \tab 407 \tab 277 \tab 19 \tab 62 \tab 70 \tab 2992 \cr
   Total \tab 1655 \tab 364 \tab 830 \tab 848 \tab 579 \tab 30 \tab 135 \tab 157 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
