% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{cathid}
\alias{cathid}
\title{If catholic what identity}
\source{
General Social Survey https://gss.norc.org
}
\description{
cathid
}
\details{
Question F. When it comes to your religious identity, would you say you are a traditional, moderate, or liberal Catholic, or do none of these describe you?
}
\section{Values}{

\itemize{
\item \code{1} traditional
\item \code{2} moderate
\item \code{3} liberal
\item \code{4} none
\item \code{6} other (specify)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5330/vshow}{GSS Data Explorer page for \code{cathid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab traditional \tab moderate \tab liberal \tab none \tab other (specify) \tab no answer \tab don't know \tab Total \cr
   1998 \tab 2432 \tab 91 \tab 98 \tab 93 \tab 48 \tab 3 \tab 67 \tab - \tab 2832 \cr
   2000 \tab 2138 \tab 197 \tab 201 \tab 158 \tab 91 \tab 4 \tab 26 \tab 2 \tab 2817 \cr
   Total \tab 4570 \tab 288 \tab 299 \tab 251 \tab 139 \tab 7 \tab 93 \tab 2 \tab 5649 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
