% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_448.R
\name{cctv}
\alias{cctv}
\title{Civil liberties - video surveillance}
\source{
General Social Survey https://gss.norc.org
}
\description{
cctv
}
\details{
Question Now I have some questions about civil liberties and public security.
Do you think that the American government should or should not have the right to do the following?
Keep people under video surveillance in public areas?  Would you say...
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should have the right
\item \verb{[2]} probably should have the right
\item \verb{[3]} probably should not have the right
\item \verb{[4]} definitely should not have the right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6315/vshow}{GSS Data Explorer page for \code{cctv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should have the right \tab probably should have the right \tab probably should not have the right \tab definitely should not have the right \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 345 \tab 534 \tab 264 \tab 215 \tab 19 \tab 13 \tab 2867 \cr
   Total \tab 1477 \tab 345 \tab 534 \tab 264 \tab 215 \tab 19 \tab 13 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutdebt}},
\code{\link{emonitor}},
\code{\link{govtinfo}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
