% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_313.R
\name{cesd2}
\alias{cesd2}
\title{How much time sleep was restless in past wk}
\source{
General Social Survey https://gss.norc.org
}
\description{
cesd2
}
\details{
Question (HAND CARD N1)
(Using this card, please tell me how much of the time during the past week...)
...your sleep was restless?
}
\section{Values}{

\itemize{
\item \code{1} none or almost none of the time
\item \code{2} some of the time
\item \code{3} most of the time
\item \code{4} all or almost all of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6298/vshow}{GSS Data Explorer page for \code{cesd2}}.
}

\seealso{
Other ESS Health: 
\code{\link{cesd1}},
\code{\link{cesd3}},
\code{\link{cesd4}},
\code{\link{cesd5}}
}
\concept{ESS Health}
\keyword{variable}
