% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_455.R
\name{change1}
\alias{change1}
\title{Help regarding important change #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
change1
}
\details{
Question Suppose you needed advice about an important change in your life -- for example, about a job or moving to another part of the country?
A. Who would you turn to first for help?
B. Who would you turn to second?
Please CIRCLE ONLY ONE FIRST CHOICE NUMBER AND ONE SECOND CHOICE NUMBER.
A. First Choice
}
\section{Values}{

\itemize{
\item \verb{[1]} spouse
\item \verb{[2]} mother
\item \verb{[3]} father
\item \verb{[4]} daughter
\item \verb{[5]} son
\item \verb{[6]} sister
\item \verb{[7]} brother
\item \verb{[8]} other relative
\item \verb{[9]} closest friend
\item \verb{[10]} other friend
\item \verb{[11]} neighbor
\item \verb{[12]} coworker
\item \verb{[13]} clergy
\item \verb{[14]} family doctor
\item \verb{[15]} prof counselor
\item \verb{[16]} lawyer
\item \verb{[17]} other
\item \verb{[18]} no one
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4179/vshow}{GSS Data Explorer page for \code{change1}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab neighbor \tab coworker \tab clergy \tab family doctor \tab prof counselor \tab lawyer \tab other \tab no one \tab don't know \tab no answer \tab Total \cr
   1986 \tab 42 \tab 688 \tab 122 \tab 58 \tab 78 \tab 63 \tab 52 \tab 35 \tab 31 \tab 153 \tab 6 \tab 7 \tab 10 \tab 32 \tab 2 \tab 13 \tab 7 \tab 10 \tab 46 \tab 2 \tab 13 \tab 1470 \cr
   Total \tab 42 \tab 688 \tab 122 \tab 58 \tab 78 \tab 63 \tab 52 \tab 35 \tab 31 \tab 153 \tab 6 \tab 7 \tab 10 \tab 32 \tab 2 \tab 13 \tab 7 \tab 10 \tab 46 \tab 2 \tab 13 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1}},
\code{\link{borrow2}},
\code{\link{change2}},
\code{\link{down1}},
\code{\link{down2}},
\code{\link{sick1}},
\code{\link{sick2}},
\code{\link{upset1}},
\code{\link{upset2}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
