% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{changed3}
\alias{changed3}
\title{Life changes from religious change3}
\source{
General Social Survey https://gss.norc.org
}
\description{
changed3
}
\details{
Question The consequences of the spiritual or religious change.
C. Third mention
}
\section{Values}{

\itemize{
\item \verb{[96]} no mention
\item \verb{[97]} nothing special happened
\item \verb{[99]} no asnwer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3076/vshow}{GSS Data Explorer page for \code{changed3}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 22 \tab 24 \tab 25 \tab 28 \tab 29 \tab 30 \tab 32 \tab 33 \tab 40 \tab 41 \tab 45 \tab 49 \tab 51 \tab 52 \tab 58 \tab 60 \tab 62 \tab 70 \tab 90 \tab 95 \tab no mention \tab no answer \tab Total \cr
   2004 \tab 2163 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 583 \tab 34 \tab 2812 \cr
   Total \tab 2163 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 583 \tab 34 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{alloflfe}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}}
}
\concept{Spirituality and Religion}
\keyword{variable}
