% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{charactr}
\alias{charactr}
\title{How likely x s situation caused by bad character}
\source{
General Social Survey https://gss.norc.org
}
\description{
charactr
}
\details{
Question In your opinion, how likely is it that NAME's situation might be caused by (READ a-f) -- very likely, somewhat likely, not very likely, or not at all likely?
A. His/her own bad character
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1988/vshow}{GSS Data Explorer page for \code{charactr}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1460 \tab 239 \tab 391 \tab 443 \tab 304 \tab 50 \tab 17 \tab - \tab - \tab 2904 \cr
   2006 \tab 3073 \tab 208 \tab 391 \tab 455 \tab 330 \tab 39 \tab 14 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 148 \tab 328 \tab 360 \tab 292 \tab 16 \tab 29 \tab - \tab - \tab 2348 \cr
   2024 \tab 1691 \tab 155 \tab 453 \tab 582 \tab 382 \tab 20 \tab 9 \tab - \tab 17 \tab 3309 \cr
   Total \tab 7399 \tab 750 \tab 1563 \tab 1840 \tab 1308 \tab 125 \tab 69 \tab 0 \tab 17 \tab 13071 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}

Other Mental Health: 
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
