% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{charisma}
\alias{charisma}
\title{Is r associated with charismatic movement}
\source{
General Social Survey https://gss.norc.org
}
\description{
charisma
}
\details{
Question Are you a charismatic or affiliated with the charismatic movement?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5331/vshow}{GSS Data Explorer page for \code{charisma}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab - \tab 149 \tab 2363 \tab 227 \tab 78 \tab 2817 \cr
   Total \tab 0 \tab 149 \tab 2363 \tab 227 \tab 78 \tab 2817 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
