% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_98.R
\name{child2}
\alias{child2}
\title{Person #2 is r's child}
\source{
General Social Survey https://gss.norc.org
}
\description{
child2
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
D2. Second Person is Child
}
\section{Values}{

\itemize{
\item \verb{[1]} mentioned
\item \verb{[2]} not mentioned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/891/vshow}{GSS Data Explorer page for \code{child2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab Total \cr
   1985 \tab 364 \tab 124 \tab 1043 \tab 3 \tab 1534 \cr
   1987 \tab 385 \tab 173 \tab 1236 \tab 25 \tab 1819 \cr
   2004 \tab 1977 \tab 93 \tab 695 \tab 47 \tab 2812 \cr
   Total \tab 2726 \tab 390 \tab 2974 \tab 75 \tab 6165 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{child1}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}}
}
\concept{Social Networks}
\keyword{variable}
