% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_98.R
\name{child3}
\alias{child3}
\title{Person #3 is r's child}
\source{
General Social Survey https://gss.norc.org
}
\description{
child3
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
D3. Third Person is Child
}
\section{Values}{

\itemize{
\item \code{1} mentioned
\item \code{2} not mentioned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/892/vshow}{GSS Data Explorer page for \code{child3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab don't know \tab Total \cr
   1985 \tab 599 \tab 98 \tab 833 \tab 4 \tab - \tab 1534 \cr
   1987 \tab 735 \tab 103 \tab 956 \tab 25 \tab - \tab 1819 \cr
   2004 \tab 2240 \tab 54 \tab 470 \tab 47 \tab 1 \tab 2812 \cr
   Total \tab 3574 \tab 255 \tab 2259 \tab 76 \tab 1 \tab 6165 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}}
}
\concept{Social Networks}
\keyword{variable}
