% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{childs_exp}
\alias{childs_exp}
\title{Total children under 18 in hh (experimental verison)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childs_exp
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDS_EXP has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDS_EXP. The original CHILDS_EXP containing data in 2022 has been renamed CHILDS_EXP_22.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0
\item \verb{[1]} 1
\item \verb{[2]} 2
\item \verb{[3]} 3
\item \verb{[4]} 4
\item \verb{[5]} 5
\item \verb{[6]} 6 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7999/vshow}{GSS Data Explorer page for \code{childs_exp}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 or more \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 1214 \tab 215 \tab 177 \tab 85 \tab 30 \tab 13 \tab 8 \tab 1794 \tab 8 \tab - \tab 3544 \cr
   2024 \tab - \tab 1145 \tab 183 \tab 155 \tab 73 \tab 29 \tab 14 \tab 5 \tab 1684 \tab 3 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 2359 \tab 398 \tab 332 \tab 158 \tab 59 \tab 27 \tab 13 \tab 3478 \tab 11 \tab 18 \tab 6853 \cr
}
}

\concept{Core}
\concept{Household}
\keyword{variable}
