% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{childs_exp_22}
\alias{childs_exp_22}
\title{Total children under 18 in hh (experimental verison)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childs_exp_22
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDS_EXP has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDS_EXP. The original CHILDS_EXP containing data in 2022 has been renamed CHILDS_EXP_22.
}
\section{Values}{

\itemize{
\item \verb{[8]} 8 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8243/vshow}{GSS Data Explorer page for \code{childs_exp_22}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 or more \tab iap \tab no answer \tab Total \cr
   2022 \tab - \tab 1214 \tab 215 \tab 177 \tab 85 \tab 30 \tab 13 \tab 4 \tab 3 \tab 1 \tab 1794 \tab 8 \tab 3544 \cr
   Total \tab 0 \tab 1214 \tab 215 \tab 177 \tab 85 \tab 30 \tab 13 \tab 4 \tab 3 \tab 1 \tab 1794 \tab 8 \tab 3544 \cr
}
}

\seealso{
Other Core: 
\code{\link{agekdbrn_9422}},
\code{\link{babies_7222}},
\code{\link{babies_exp_22}},
\code{\link{childsinhh_2122}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp_22}},
\code{\link{teens_7222}},
\code{\link{teens_exp_22}},
\code{\link{unrelat_7222}}

Other Household: 
\code{\link{babies_7222}},
\code{\link{babies_exp_22}},
\code{\link{childsinhh_2122}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp_22}},
\code{\link{teens_7222}},
\code{\link{teens_exp_22}},
\code{\link{unrelat_7222}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
