% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{childsinhh}
\alias{childsinhh}
\title{Total children in hh (household composition version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childsinhh
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDSINHH has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDSINHH. The original CHILDSINHH containing data from 2021 to 2022 has been renamed CHILDSINHH_2122.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0
\item \verb{[1]} 1
\item \verb{[2]} 2
\item \verb{[3]} 3
\item \verb{[4]} 4
\item \verb{[5]} 5
\item \verb{[6]} 6 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8006/vshow}{GSS Data Explorer page for \code{childsinhh}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 or more \tab iap \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab - \tab 2845 \tab 444 \tab 347 \tab 120 \tab 47 \tab 13 \tab 10 \tab 203 \tab 3 \tab - \tab 4032 \cr
   2022 \tab - \tab 1342 \tab 158 \tab 167 \tab 75 \tab 15 \tab 6 \tab 5 \tab 1764 \tab 4 \tab 8 \tab 3544 \cr
   2024 \tab - \tab 1244 \tab 169 \tab 146 \tab 71 \tab 31 \tab 10 \tab 5 \tab 1630 \tab 2 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 5431 \tab 771 \tab 660 \tab 266 \tab 93 \tab 29 \tab 20 \tab 3597 \tab 9 \tab 9 \tab 10885 \cr
}
}

\concept{Core}
\concept{Household}
\keyword{variable}
