% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{childsinhh_2122}
\alias{childsinhh_2122}
\title{Total children in hh (household composition version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childsinhh_2122
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDSINHH has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDSINHH. The original CHILDSINHH containing data from 2021 to 2022 has been renamed CHILDSINHH_2122.
}
\section{Values}{

\itemize{
\item \verb{[8]} 8 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8242/vshow}{GSS Data Explorer page for \code{childsinhh_2122}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab iap \tab skipped on web \tab 8 or more \tab no answer \tab Total \cr
   2021 \tab - \tab 2845 \tab 444 \tab 347 \tab 120 \tab 47 \tab 13 \tab 8 \tab 2 \tab 203 \tab 3 \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 1342 \tab 158 \tab 167 \tab 75 \tab 15 \tab 6 \tab 2 \tab 1 \tab 1764 \tab 4 \tab 2 \tab 8 \tab 3544 \cr
   Total \tab 0 \tab 4187 \tab 602 \tab 514 \tab 195 \tab 62 \tab 19 \tab 10 \tab 3 \tab 1967 \tab 7 \tab 2 \tab 8 \tab 7576 \cr
}
}

\seealso{
Other Core: 
\code{\link{agekdbrn_9422}},
\code{\link{babies_7222}},
\code{\link{babies_exp_22}},
\code{\link{childs_exp_22}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp_22}},
\code{\link{teens_7222}},
\code{\link{teens_exp_22}},
\code{\link{unrelat_7222}}

Other Household: 
\code{\link{babies_7222}},
\code{\link{babies_exp_22}},
\code{\link{childs_exp_22}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp_22}},
\code{\link{teens_7222}},
\code{\link{teens_exp_22}},
\code{\link{unrelat_7222}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
