% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{chkspfin}
\alias{chkspfin}
\title{Government should verify spouse's financial assets}
\source{
General Social Survey https://gss.norc.org
}
\description{
chkspfin
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should contact other people and verify information provided by the individual concerning his or her: (READ A-C)
B. Spouse's financial assets and liabilities
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5224/vshow}{GSS Data Explorer page for \code{chkspfin}}.
}

\seealso{
Other National Security: 
\code{\link{chkfinan}},
\code{\link{chktaxes}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{punarms}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}},
\code{\link{taketrck}}

Other Federal Government: 
\code{\link{chkfinan}},
\code{\link{chktaxes}},
\code{\link{knomentl}}
}
\concept{Federal Government}
\concept{National Security}
\keyword{variable}
