% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{chldadlt_blk}
\alias{chldadlt_blk}
\title{Percent of adults you knew (friends, care takers, neighbors) during childhood who were black}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldadlt_blk
}
\details{
Question Not including your family, what percentage of the adults you knew (friends, care takers, neighbors) belonged to each of the following categories? - Black
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8072/vshow}{GSS Data Explorer page for \code{chldadlt_blk}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 24 \tab 25 \tab 26 \tab 28 \tab 3 \tab 30 \tab 32 \tab 33 \tab 35 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 5 \tab 50 \tab 53 \tab 55 \tab 56 \tab 57 \tab 6 \tab 60 \tab 61 \tab 64 \tab 65 \tab 66 \tab 7 \tab 70 \tab 75 \tab 77 \tab 78 \tab 8 \tab 80 \tab 82 \tab 83 \tab 85 \tab 89 \tab 9 \tab 90 \tab 92 \tab 94 \tab 95 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 390 \tab 41 \tab 169 \tab 66 \tab 5 \tab 2 \tab 3 \tab 1 \tab 44 \tab 1 \tab 1 \tab 2 \tab 2 \tab 38 \tab 139 \tab 1 \tab 2 \tab 1 \tab 70 \tab 1 \tab 1 \tab 11 \tab 60 \tab 1 \tab 6 \tab 3 \tab 2 \tab 2 \tab 2 \tab 4 \tab 29 \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 155 \tab 61 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 12 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 24 \tab 13 \tab 2 \tab 1 \tab 3 \tab 27 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 21 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab 35 \tab 1665 \tab 24 \tab 109 \tab 3309 \cr
   Total \tab 0 \tab 390 \tab 41 \tab 169 \tab 66 \tab 5 \tab 2 \tab 3 \tab 1 \tab 44 \tab 1 \tab 1 \tab 2 \tab 2 \tab 38 \tab 139 \tab 1 \tab 2 \tab 1 \tab 70 \tab 1 \tab 1 \tab 11 \tab 60 \tab 1 \tab 6 \tab 3 \tab 2 \tab 2 \tab 2 \tab 4 \tab 29 \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 155 \tab 61 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 12 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 24 \tab 13 \tab 2 \tab 1 \tab 3 \tab 27 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 21 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab 35 \tab 1665 \tab 24 \tab 109 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}}
}
\concept{Mental Health}
\keyword{variable}
