% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{chldadlt_hisp}
\alias{chldadlt_hisp}
\title{Percent of adults you knew (friends, care takers, neighbors) during childhood who were latino}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldadlt_hisp
}
\details{
Question Not including your family, what percentage of the adults you knew (friends, care takers, neighbors) belonged to each of the following categories? - Latino
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8073/vshow}{GSS Data Explorer page for \code{chldadlt_hisp}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 13 \tab 15 \tab 16 \tab 17 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 25 \tab 26 \tab 28 \tab 3 \tab 30 \tab 33 \tab 34 \tab 35 \tab 37 \tab 4 \tab 40 \tab 44 \tab 45 \tab 5 \tab 50 \tab 55 \tab 56 \tab 6 \tab 60 \tab 61 \tab 66 \tab 7 \tab 70 \tab 75 \tab 76 \tab 8 \tab 80 \tab 85 \tab 90 \tab 94 \tab 95 \tab 98 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 597 \tab 39 \tab 163 \tab 41 \tab 4 \tab 2 \tab 2 \tab 64 \tab 1 \tab 2 \tab 3 \tab 27 \tab 124 \tab 3 \tab 2 \tab 55 \tab 1 \tab 3 \tab 10 \tab 41 \tab 6 \tab 1 \tab 4 \tab 1 \tab 6 \tab 26 \tab 1 \tab 1 \tab 142 \tab 32 \tab 2 \tab 1 \tab 6 \tab 8 \tab 1 \tab 2 \tab 3 \tab 10 \tab 6 \tab 1 \tab 2 \tab 14 \tab 2 \tab 10 \tab 1 \tab 1 \tab 2 \tab 35 \tab 1665 \tab 24 \tab 109 \tab 3309 \cr
   Total \tab 0 \tab 597 \tab 39 \tab 163 \tab 41 \tab 4 \tab 2 \tab 2 \tab 64 \tab 1 \tab 2 \tab 3 \tab 27 \tab 124 \tab 3 \tab 2 \tab 55 \tab 1 \tab 3 \tab 10 \tab 41 \tab 6 \tab 1 \tab 4 \tab 1 \tab 6 \tab 26 \tab 1 \tab 1 \tab 142 \tab 32 \tab 2 \tab 1 \tab 6 \tab 8 \tab 1 \tab 2 \tab 3 \tab 10 \tab 6 \tab 1 \tab 2 \tab 14 \tab 2 \tab 10 \tab 1 \tab 1 \tab 2 \tab 35 \tab 1665 \tab 24 \tab 109 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}}
}
\concept{Mental Health}
\keyword{variable}
