% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{chldcare}
\alias{chldcare}
\title{Should families get benefits if both parents work}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldcare
}
\details{
Question Do you agree or disagree . . .
B. Families should receive financial benefits for child care when both parents work.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4413/vshow}{GSS Data Explorer page for \code{chldcare}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 204 \tab 407 \tab 261 \tab 389 \tab 71 \tab 35 \tab 80 \tab 2992 \cr
   2002 \tab 1594 \tab 333 \tab 312 \tab 206 \tab 216 \tab 74 \tab 29 \tab 1 \tab 2765 \cr
   Total \tab 3139 \tab 537 \tab 719 \tab 467 \tab 605 \tab 145 \tab 64 \tab 81 \tab 5757 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
