% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_687.R
\name{chldchld_blk}
\alias{chldchld_blk}
\title{Percent of children you knew (friends, classmates) during childhood who were black}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldchld_blk
}
\details{
Question Not including your family, what percentage of the children you knew (friends, classmates belonged to each of the following categories? - Black
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8077/vshow}{GSS Data Explorer page for \code{chldchld_blk}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 22 \tab 25 \tab 28 \tab 29 \tab 3 \tab 30 \tab 33 \tab 35 \tab 36 \tab 37 \tab 4 \tab 40 \tab 42 \tab 45 \tab 46 \tab 48 \tab 49 \tab 5 \tab 50 \tab 52 \tab 54 \tab 55 \tab 56 \tab 6 \tab 60 \tab 63 \tab 64 \tab 65 \tab 7 \tab 70 \tab 75 \tab 78 \tab 8 \tab 80 \tab 84 \tab 85 \tab 87 \tab 89 \tab 9 \tab 90 \tab 91 \tab 92 \tab 94 \tab 95 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 355 \tab 58 \tab 168 \tab 56 \tab 5 \tab 49 \tab 1 \tab 1 \tab 3 \tab 2 \tab 52 \tab 130 \tab 3 \tab 76 \tab 1 \tab 1 \tab 14 \tab 49 \tab 8 \tab 7 \tab 1 \tab 1 \tab 8 \tab 39 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 156 \tab 71 \tab 2 \tab 1 \tab 1 \tab 4 \tab 3 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 25 \tab 7 \tab 1 \tab 5 \tab 25 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 20 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 4 \tab 39 \tab 1666 \tab 26 \tab 114 \tab 3309 \cr
   Total \tab 0 \tab 355 \tab 58 \tab 168 \tab 56 \tab 5 \tab 49 \tab 1 \tab 1 \tab 3 \tab 2 \tab 52 \tab 130 \tab 3 \tab 76 \tab 1 \tab 1 \tab 14 \tab 49 \tab 8 \tab 7 \tab 1 \tab 1 \tab 8 \tab 39 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 156 \tab 71 \tab 2 \tab 1 \tab 1 \tab 4 \tab 3 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 25 \tab 7 \tab 1 \tab 5 \tab 25 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 20 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 4 \tab 39 \tab 1666 \tab 26 \tab 114 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_oth}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_wht}}
}
\concept{Mental Health}
\keyword{variable}
