% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{chldchld_oth}
\alias{chldchld_oth}
\title{Percent of children you knew (friends, classmates) during childhood who were some other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldchld_oth
}
\details{
Question Not including your family, what percentage of the children you knew (friends, classmates belonged to each of the following categories? - Some other race
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8079/vshow}{GSS Data Explorer page for \code{chldchld_oth}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 15 \tab 18 \tab 19 \tab 2 \tab 20 \tab 22 \tab 25 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 33 \tab 35 \tab 4 \tab 40 \tab 45 \tab 47 \tab 5 \tab 50 \tab 51 \tab 6 \tab 60 \tab 7 \tab 70 \tab 8 \tab 80 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1168 \tab 19 \tab 63 \tab 7 \tab 2 \tab 3 \tab 8 \tab 1 \tab 2 \tab 15 \tab 35 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 6 \tab 7 \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 90 \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 3 \tab 2 \tab 39 \tab 1666 \tab 26 \tab 114 \tab 3309 \cr
   Total \tab 0 \tab 1168 \tab 19 \tab 63 \tab 7 \tab 2 \tab 3 \tab 8 \tab 1 \tab 2 \tab 15 \tab 35 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 6 \tab 7 \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 90 \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 3 \tab 2 \tab 39 \tab 1666 \tab 26 \tab 114 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_wht}}
}
\concept{Mental Health}
\keyword{variable}
