% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{chldchld_wht}
\alias{chldchld_wht}
\title{Percent of children you knew (friends, classmates) during childhood who were white}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldchld_wht
}
\details{
Question Not including your family, what percentage of the children you knew (friends, classmates belonged to each of the following categories? - White
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8080/vshow}{GSS Data Explorer page for \code{chldchld_wht}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 13 \tab 15 \tab 17 \tab 2 \tab 20 \tab 22 \tab 25 \tab 28 \tab 29 \tab 3 \tab 30 \tab 33 \tab 34 \tab 35 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 55 \tab 56 \tab 57 \tab 6 \tab 60 \tab 61 \tab 63 \tab 64 \tab 65 \tab 67 \tab 69 \tab 70 \tab 73 \tab 75 \tab 76 \tab 78 \tab 79 \tab 8 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 9 \tab 90 \tab 91 \tab 92 \tab 93 \tab 94 \tab 95 \tab 96 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 160 \tab 8 \tab 69 \tab 171 \tab 3 \tab 2 \tab 18 \tab 1 \tab 4 \tab 65 \tab 3 \tab 52 \tab 1 \tab 3 \tab 2 \tab 45 \tab 11 \tab 3 \tab 6 \tab 2 \tab 5 \tab 2 \tab 2 \tab 63 \tab 1 \tab 24 \tab 1 \tab 1 \tab 1 \tab 2 \tab 22 \tab 118 \tab 11 \tab 3 \tab 1 \tab 2 \tab 63 \tab 1 \tab 3 \tab 2 \tab 26 \tab 2 \tab 1 \tab 66 \tab 2 \tab 45 \tab 3 \tab 3 \tab 1 \tab 1 \tab 74 \tab 1 \tab 1 \tab 4 \tab 3 \tab 36 \tab 1 \tab 2 \tab 9 \tab 6 \tab 1 \tab 56 \tab 1 \tab 4 \tab 9 \tab 12 \tab 45 \tab 7 \tab 18 \tab 39 \tab 28 \tab 39 \tab 1666 \tab 26 \tab 114 \tab 3309 \cr
   Total \tab 0 \tab 160 \tab 8 \tab 69 \tab 171 \tab 3 \tab 2 \tab 18 \tab 1 \tab 4 \tab 65 \tab 3 \tab 52 \tab 1 \tab 3 \tab 2 \tab 45 \tab 11 \tab 3 \tab 6 \tab 2 \tab 5 \tab 2 \tab 2 \tab 63 \tab 1 \tab 24 \tab 1 \tab 1 \tab 1 \tab 2 \tab 22 \tab 118 \tab 11 \tab 3 \tab 1 \tab 2 \tab 63 \tab 1 \tab 3 \tab 2 \tab 26 \tab 2 \tab 1 \tab 66 \tab 2 \tab 45 \tab 3 \tab 3 \tab 1 \tab 1 \tab 74 \tab 1 \tab 1 \tab 4 \tab 3 \tab 36 \tab 1 \tab 2 \tab 9 \tab 6 \tab 1 \tab 56 \tab 1 \tab 4 \tab 9 \tab 12 \tab 45 \tab 7 \tab 18 \tab 39 \tab 28 \tab 39 \tab 1666 \tab 26 \tab 114 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}}
}
\concept{Mental Health}
\keyword{variable}
