% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_558.R
\name{chldeduc}
\alias{chldeduc}
\title{Education improving?}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldeduc
}
\details{
Question Compared with 10 years ago, would you say that American children today are much better off, better off, about the same, worse off, or much worse off regarding:
A. The quality of their education.
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} better
\item \verb{[3]} about the same
\item \verb{[4]} worse
\item \verb{[5]} much worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5108/vshow}{GSS Data Explorer page for \code{chldeduc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much better \tab better \tab about the same \tab worse \tab much worse \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 206 \tab 465 \tab 302 \tab 319 \tab 44 \tab 23 \tab 13 \tab 1372 \cr
   Total \tab 0 \tab 206 \tab 465 \tab 302 \tab 319 \tab 44 \tab 23 \tab 13 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldlove}},
\code{\link{chldmorl}},
\code{\link{chldsafe}},
\code{\link{chldsup}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolepa}},
\code{\link{roletchr}}
}
\concept{Children}
\keyword{variable}
