% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{chldfam_blk}
\alias{chldfam_blk}
\title{Percent of family during childhood who were black}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldfam_blk
}
\details{
Question What percentage of your family belong to each of the following categories? - Black
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8082/vshow}{GSS Data Explorer page for \code{chldfam_blk}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 13 \tab 14 \tab 15 \tab 2 \tab 20 \tab 25 \tab 30 \tab 33 \tab 4 \tab 40 \tab 42 \tab 5 \tab 50 \tab 51 \tab 6 \tab 60 \tab 65 \tab 70 \tab 71 \tab 74 \tab 75 \tab 78 \tab 8 \tab 80 \tab 83 \tab 85 \tab 89 \tab 90 \tab 94 \tab 95 \tab 96 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1115 \tab 17 \tab 19 \tab 145 \tab 4 \tab 1 \tab 1 \tab 2 \tab 9 \tab 22 \tab 3 \tab 8 \tab 1 \tab 2 \tab 9 \tab 1 \tab 16 \tab 34 \tab 1 \tab 3 \tab 8 \tab 1 \tab 8 \tab 3 \tab 2 \tab 6 \tab 1 \tab 1 \tab 20 \tab 1 \tab 5 \tab 3 \tab 23 \tab 1 \tab 10 \tab 2 \tab 3 \tab 5 \tab 2 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
   Total \tab 0 \tab 1115 \tab 17 \tab 19 \tab 145 \tab 4 \tab 1 \tab 1 \tab 2 \tab 9 \tab 22 \tab 3 \tab 8 \tab 1 \tab 2 \tab 9 \tab 1 \tab 16 \tab 34 \tab 1 \tab 3 \tab 8 \tab 1 \tab 8 \tab 3 \tab 2 \tab 6 \tab 1 \tab 1 \tab 20 \tab 1 \tab 5 \tab 3 \tab 23 \tab 1 \tab 10 \tab 2 \tab 3 \tab 5 \tab 2 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}}
}
\concept{Mental Health}
\keyword{variable}
