% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{chldfam_hisp}
\alias{chldfam_hisp}
\title{Percent of family during childhood who were latino}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldfam_hisp
}
\details{
Question What percentage of your family belong to each of the following categories? - Latino
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8083/vshow}{GSS Data Explorer page for \code{chldfam_hisp}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 15 \tab 19 \tab 2 \tab 20 \tab 22 \tab 24 \tab 25 \tab 3 \tab 30 \tab 35 \tab 38 \tab 4 \tab 40 \tab 45 \tab 5 \tab 50 \tab 53 \tab 6 \tab 60 \tab 65 \tab 70 \tab 71 \tab 75 \tab 77 \tab 80 \tab 85 \tab 90 \tab 91 \tab 94 \tab 95 \tab 96 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1221 \tab 13 \tab 21 \tab 102 \tab 2 \tab 5 \tab 1 \tab 10 \tab 21 \tab 2 \tab 1 \tab 7 \tab 2 \tab 7 \tab 3 \tab 1 \tab 1 \tab 8 \tab 2 \tab 27 \tab 31 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 7 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
   Total \tab 0 \tab 1221 \tab 13 \tab 21 \tab 102 \tab 2 \tab 5 \tab 1 \tab 10 \tab 21 \tab 2 \tab 1 \tab 7 \tab 2 \tab 7 \tab 3 \tab 1 \tab 1 \tab 8 \tab 2 \tab 27 \tab 31 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 7 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}}
}
\concept{Mental Health}
\keyword{variable}
