% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{chldfam_oth}
\alias{chldfam_oth}
\title{Percent of family during childhood who were some other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldfam_oth
}
\details{
Question What percentage of your family belong to each of the following categories? - Some other race
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8084/vshow}{GSS Data Explorer page for \code{chldfam_oth}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 13 \tab 14 \tab 15 \tab 18 \tab 2 \tab 20 \tab 22 \tab 25 \tab 29 \tab 3 \tab 30 \tab 31 \tab 4 \tab 40 \tab 5 \tab 50 \tab 51 \tab 53 \tab 56 \tab 6 \tab 60 \tab 7 \tab 75 \tab 8 \tab 80 \tab 96 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1323 \tab 18 \tab 40 \tab 14 \tab 1 \tab 3 \tab 2 \tab 8 \tab 1 \tab 5 \tab 16 \tab 2 \tab 13 \tab 1 \tab 4 \tab 10 \tab 1 \tab 2 \tab 2 \tab 24 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
   Total \tab 0 \tab 1323 \tab 18 \tab 40 \tab 14 \tab 1 \tab 3 \tab 2 \tab 8 \tab 1 \tab 5 \tab 16 \tab 2 \tab 13 \tab 1 \tab 4 \tab 10 \tab 1 \tab 2 \tab 2 \tab 24 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_wht}}
}
\concept{Mental Health}
\keyword{variable}
