% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{chldfam_wht}
\alias{chldfam_wht}
\title{Percent of family during childhood who were white}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldfam_wht
}
\details{
Question What percentage of your family belong to each of the following categories? - White
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8085/vshow}{GSS Data Explorer page for \code{chldfam_wht}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 14 \tab 15 \tab 17 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 25 \tab 3 \tab 30 \tab 36 \tab 38 \tab 4 \tab 40 \tab 45 \tab 5 \tab 50 \tab 6 \tab 60 \tab 62 \tab 65 \tab 67 \tab 7 \tab 70 \tab 75 \tab 76 \tab 78 \tab 80 \tab 81 \tab 82 \tab 83 \tab 85 \tab 87 \tab 89 \tab 9 \tab 90 \tab 92 \tab 93 \tab 94 \tab 95 \tab 96 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 367 \tab 12 \tab 24 \tab 782 \tab 1 \tab 3 \tab 1 \tab 6 \tab 1 \tab 1 \tab 4 \tab 25 \tab 1 \tab 1 \tab 16 \tab 1 \tab 12 \tab 1 \tab 2 \tab 2 \tab 14 \tab 1 \tab 26 \tab 52 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 11 \tab 13 \tab 1 \tab 2 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 33 \tab 1 \tab 1 \tab 2 \tab 18 \tab 1 \tab 5 \tab 12 \tab 21 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
   Total \tab 0 \tab 367 \tab 12 \tab 24 \tab 782 \tab 1 \tab 3 \tab 1 \tab 6 \tab 1 \tab 1 \tab 4 \tab 25 \tab 1 \tab 1 \tab 16 \tab 1 \tab 12 \tab 1 \tab 2 \tab 2 \tab 14 \tab 1 \tab 26 \tab 52 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 11 \tab 13 \tab 1 \tab 2 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 33 \tab 1 \tab 1 \tab 2 \tab 18 \tab 1 \tab 5 \tab 12 \tab 21 \tab 18 \tab 1664 \tab 16 \tab 93 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}}
}
\concept{Mental Health}
\keyword{variable}
