% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_329.R
\name{chldfrnd}
\alias{chldfrnd}
\title{How willing r have children make friends with ch}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldfrnd
}
\details{
Question C. How willing would you be to have your children make friends with (John/Mary)?
}
\section{Values}{

\itemize{
\item \code{1} definitely willing
\item \code{2} probably willing
\item \code{3} probably unwilling
\item \code{4} definitely unwilling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2935/vshow}{GSS Data Explorer page for \code{chldfrnd}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldsch}},
\code{\link{spendeve}}
}
\concept{Mental Health II}
\keyword{variable}
