% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{chldhlth}
\alias{chldhlth}
\title{Health care improving?}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldhlth
}
\details{
Question Compared with 10 years ago, would you say that American children today are much better off, better off, about the same, worse off, or much worse off regarding:
G. Getting health care when they need it.
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} better
\item \code{3} about the same
\item \code{4} worse
\item \code{5} much worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5114/vshow}{GSS Data Explorer page for \code{chldhlth}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much better \tab better \tab about the same \tab worse \tab much worse \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 101 \tab 510 \tab 415 \tab 278 \tab 25 \tab 28 \tab 15 \tab 1372 \cr
   Total \tab 0 \tab 101 \tab 510 \tab 415 \tab 278 \tab 25 \tab 28 \tab 15 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldhome}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{sppregnt}}
}
\concept{Children}
\keyword{variable}
