% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_180.R
\name{chldhrt}
\alias{chldhrt}
\title{Children hurt from traditional family}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldhrt
}
\details{
Question Who is hurt by this arrangement: men, women, children, everyone, or no one? PROBE:  Anyone else? UNTIL R SAYS NO.
C. Children
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1685/vshow}{GSS Data Explorer page for \code{chldhrt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 197 \tab 780 \tab 82 \tab 401 \tab 2904 \cr
   Total \tab 1444 \tab 197 \tab 780 \tab 82 \tab 401 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allhrt}},
\code{\link{fekids1}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{womenhrt}}
}
\concept{Gender}
\keyword{variable}
