% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{chldpref_next}
\alias{chldpref_next}
\title{R ideal number of children - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldpref_next
}
\details{
Question For you personally, what is the ideal number of children you would like to have or would have liked to have had?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[7]} 7 or more
\item \verb{[8]} as many as you want
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8257/vshow}{GSS Data Explorer page for \code{chldpref_next}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 or more \tab as many as you want \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 127 \tab 85 \tab 432 \tab 203 \tab 100 \tab 31 \tab 15 \tab 12 \tab 58 \tab 2244 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 127 \tab 85 \tab 432 \tab 203 \tab 100 \tab 31 \tab 15 \tab 12 \tab 58 \tab 2244 \tab 2 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{happy_next}},
\code{\link{knwsexor}},
\code{\link{perla_self}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other Children: 
\code{\link{chldidel_next}}
}
\concept{Children}
\concept{GSS Next}
\keyword{variable}
