% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chldprob}
\alias{chldprob}
\title{How serious r consider his or her problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldprob
}
\details{
Question How serious would you consider (John/Mary)'s problem, if any, to be: very serious, somewhat serious, not very serious, or not at all serious?
}
\section{Values}{

\itemize{
\item \verb{[1]} very serious
\item \verb{[2]} somewhat serious
\item \verb{[3]} not very serious
\item \verb{[4]} not at all serious
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2919/vshow}{GSS Data Explorer page for \code{chldprob}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very serious \tab somewhat serious \tab not very serious \tab not at all serious \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2002 \tab 1372 \tab 625 \tab 389 \tab 168 \tab 171 \tab 28 \tab 12 \tab - \tab - \tab 2765 \cr
   2024 \tab 1659 \tab 652 \tab 616 \tab 258 \tab 92 \tab 10 \tab 4 \tab - \tab 18 \tab 3309 \cr
   Total \tab 3031 \tab 1277 \tab 1005 \tab 426 \tab 263 \tab 38 \tab 16 \tab 0 \tab 18 \tab 6074 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldvig}},
\code{\link{stressfl}}

Other Children: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldvig}},
\code{\link{stressfl}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
