% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{chldsex1}
\alias{chldsex1}
\title{Qualities for boys or girls:variant}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldsex1
}
\details{
Question B. When you rated the importance of various qualities for children, were you thinking mostly about boys or mostly about girls?
}
\section{Values}{

\itemize{
\item \code{1} mostly boys
\item \code{2} mostly girls
\item \code{3} both equally
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/519/vshow}{GSS Data Explorer page for \code{chldsex1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab mostly boys \tab mostly girls \tab both equally \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 359 \tab 154 \tab 895 \tab 16 \tab 49 \tab 1473 \cr
   Total \tab 0 \tab 359 \tab 154 \tab 895 \tab 16 \tab 49 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{chldsex}},
\code{\link{helpoth}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socrel}},
\code{\link{thnkself}},
\code{\link{workhard}},
\code{\link{youngen}}
}
\concept{Core}
\keyword{variable}
