% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_687.R
\name{chldshop_blk}
\alias{chldshop_blk}
\title{Percent of people who shopped at the supermarket during childhood who were black}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldshop_blk
}
\details{
Question Think about the supermarket where your family shopped. What percentage of the people who shopped at the supermarket belong to each of the following categories? - Black
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8087/vshow}{GSS Data Explorer page for \code{chldshop_blk}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 35 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 43 \tab 44 \tab 45 \tab 49 \tab 5 \tab 50 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 6 \tab 60 \tab 63 \tab 67 \tab 7 \tab 70 \tab 71 \tab 75 \tab 77 \tab 8 \tab 80 \tab 81 \tab 84 \tab 85 \tab 9 \tab 90 \tab 91 \tab 94 \tab 95 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 315 \tab 35 \tab 175 \tab 31 \tab 6 \tab 4 \tab 1 \tab 49 \tab 1 \tab 2 \tab 3 \tab 39 \tab 149 \tab 1 \tab 5 \tab 1 \tab 1 \tab 80 \tab 1 \tab 3 \tab 15 \tab 80 \tab 1 \tab 3 \tab 7 \tab 13 \tab 2 \tab 2 \tab 1 \tab 5 \tab 53 \tab 2 \tab 2 \tab 7 \tab 1 \tab 146 \tab 88 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 4 \tab 11 \tab 1 \tab 2 \tab 3 \tab 13 \tab 1 \tab 10 \tab 1 \tab 4 \tab 18 \tab 1 \tab 1 \tab 3 \tab 5 \tab 12 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
   Total \tab 0 \tab 315 \tab 35 \tab 175 \tab 31 \tab 6 \tab 4 \tab 1 \tab 49 \tab 1 \tab 2 \tab 3 \tab 39 \tab 149 \tab 1 \tab 5 \tab 1 \tab 1 \tab 80 \tab 1 \tab 3 \tab 15 \tab 80 \tab 1 \tab 3 \tab 7 \tab 13 \tab 2 \tab 2 \tab 1 \tab 5 \tab 53 \tab 2 \tab 2 \tab 7 \tab 1 \tab 146 \tab 88 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 4 \tab 11 \tab 1 \tab 2 \tab 3 \tab 13 \tab 1 \tab 10 \tab 1 \tab 4 \tab 18 \tab 1 \tab 1 \tab 3 \tab 5 \tab 12 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_oth}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_wht}}
}
\concept{Mental Health}
\keyword{variable}
