% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_687.R
\name{chldshop_hisp}
\alias{chldshop_hisp}
\title{Percent of people who shopped at the supermarket during childhood who were latino}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldshop_hisp
}
\details{
Question Think about the supermarket where your family shopped. What percentage of the people who shopped at the supermarket belong to each of the following categories? - Latino
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8088/vshow}{GSS Data Explorer page for \code{chldshop_hisp}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 33 \tab 34 \tab 35 \tab 4 \tab 40 \tab 42 \tab 43 \tab 45 \tab 47 \tab 49 \tab 5 \tab 50 \tab 53 \tab 6 \tab 60 \tab 67 \tab 7 \tab 70 \tab 75 \tab 8 \tab 80 \tab 85 \tab 89 \tab 9 \tab 90 \tab 92 \tab 95 \tab 98 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 474 \tab 24 \tab 188 \tab 26 \tab 7 \tab 2 \tab 1 \tab 64 \tab 1 \tab 1 \tab 2 \tab 2 \tab 40 \tab 154 \tab 3 \tab 4 \tab 1 \tab 2 \tab 71 \tab 2 \tab 2 \tab 2 \tab 13 \tab 55 \tab 7 \tab 2 \tab 14 \tab 4 \tab 22 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 150 \tab 32 \tab 1 \tab 8 \tab 10 \tab 1 \tab 1 \tab 7 \tab 5 \tab 4 \tab 8 \tab 2 \tab 1 \tab 4 \tab 9 \tab 1 \tab 1 \tab 1 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
   Total \tab 0 \tab 474 \tab 24 \tab 188 \tab 26 \tab 7 \tab 2 \tab 1 \tab 64 \tab 1 \tab 1 \tab 2 \tab 2 \tab 40 \tab 154 \tab 3 \tab 4 \tab 1 \tab 2 \tab 71 \tab 2 \tab 2 \tab 2 \tab 13 \tab 55 \tab 7 \tab 2 \tab 14 \tab 4 \tab 22 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 150 \tab 32 \tab 1 \tab 8 \tab 10 \tab 1 \tab 1 \tab 7 \tab 5 \tab 4 \tab 8 \tab 2 \tab 1 \tab 4 \tab 9 \tab 1 \tab 1 \tab 1 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_oth}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_wht}}
}
\concept{Mental Health}
\keyword{variable}
