% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_687.R
\name{chldshop_oth}
\alias{chldshop_oth}
\title{Percent of people who shopped at the supermarket during childhood who were some other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldshop_oth
}
\details{
Question Think about the supermarket where your family shopped. What percentage of the people who shopped at the supermarket belong to each of the following categories? - Some other race
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8089/vshow}{GSS Data Explorer page for \code{chldshop_oth}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 13 \tab 15 \tab 17 \tab 2 \tab 20 \tab 22 \tab 24 \tab 25 \tab 26 \tab 29 \tab 3 \tab 30 \tab 33 \tab 35 \tab 4 \tab 5 \tab 50 \tab 56 \tab 6 \tab 60 \tab 7 \tab 70 \tab 8 \tab 80 \tab 9 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1133 \tab 19 \tab 56 \tab 8 \tab 6 \tab 2 \tab 3 \tab 13 \tab 1 \tab 13 \tab 52 \tab 2 \tab 1 \tab 10 \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab 2 \tab 6 \tab 83 \tab 4 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
   Total \tab 0 \tab 1133 \tab 19 \tab 56 \tab 8 \tab 6 \tab 2 \tab 3 \tab 13 \tab 1 \tab 13 \tab 52 \tab 2 \tab 1 \tab 10 \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab 2 \tab 6 \tab 83 \tab 4 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_wht}}
}
\concept{Mental Health}
\keyword{variable}
