% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_687.R
\name{chldshop_wht}
\alias{chldshop_wht}
\title{Percent of people who shopped at the supermarket during childhood who were white}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldshop_wht
}
\details{
Question Think about the supermarket where your family shopped. What percentage of the people who shopped at the supermarket belong to each of the following categories? - White
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8090/vshow}{GSS Data Explorer page for \code{chldshop_wht}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 14 \tab 15 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 28 \tab 29 \tab 3 \tab 30 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 4 \tab 40 \tab 42 \tab 43 \tab 44 \tab 45 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 63 \tab 65 \tab 67 \tab 68 \tab 69 \tab 7 \tab 70 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 79 \tab 80 \tab 81 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 9 \tab 90 \tab 91 \tab 92 \tab 93 \tab 94 \tab 95 \tab 96 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 127 \tab 3 \tab 49 \tab 142 \tab 2 \tab 2 \tab 9 \tab 1 \tab 1 \tab 80 \tab 1 \tab 6 \tab 1 \tab 2 \tab 64 \tab 2 \tab 1 \tab 3 \tab 4 \tab 42 \tab 1 \tab 10 \tab 1 \tab 10 \tab 1 \tab 1 \tab 2 \tab 1 \tab 89 \tab 1 \tab 1 \tab 5 \tab 26 \tab 2 \tab 2 \tab 1 \tab 20 \tab 145 \tab 1 \tab 2 \tab 15 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 66 \tab 1 \tab 5 \tab 19 \tab 1 \tab 2 \tab 1 \tab 1 \tab 75 \tab 1 \tab 2 \tab 2 \tab 44 \tab 1 \tab 1 \tab 4 \tab 1 \tab 80 \tab 2 \tab 1 \tab 1 \tab 33 \tab 2 \tab 2 \tab 5 \tab 6 \tab 3 \tab 69 \tab 2 \tab 8 \tab 5 \tab 9 \tab 40 \tab 12 \tab 9 \tab 19 \tab 17 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
   Total \tab 0 \tab 127 \tab 3 \tab 49 \tab 142 \tab 2 \tab 2 \tab 9 \tab 1 \tab 1 \tab 80 \tab 1 \tab 6 \tab 1 \tab 2 \tab 64 \tab 2 \tab 1 \tab 3 \tab 4 \tab 42 \tab 1 \tab 10 \tab 1 \tab 10 \tab 1 \tab 1 \tab 2 \tab 1 \tab 89 \tab 1 \tab 1 \tab 5 \tab 26 \tab 2 \tab 2 \tab 1 \tab 20 \tab 145 \tab 1 \tab 2 \tab 15 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 66 \tab 1 \tab 5 \tab 19 \tab 1 \tab 2 \tab 1 \tab 1 \tab 75 \tab 1 \tab 2 \tab 2 \tab 44 \tab 1 \tab 1 \tab 4 \tab 1 \tab 80 \tab 2 \tab 1 \tab 1 \tab 33 \tab 2 \tab 2 \tab 5 \tab 6 \tab 3 \tab 69 \tab 2 \tab 8 \tab 5 \tab 9 \tab 40 \tab 12 \tab 9 \tab 19 \tab 17 \tab 67 \tab 1667 \tab 26 \tab 108 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_oth}}
}
\concept{Mental Health}
\keyword{variable}
