% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{chldskls}
\alias{chldskls}
\title{Acquiring job skills improving?}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldskls
}
\details{
Question Compared with 10 years ago, would you say that American children today are much better off, better off, about the same, worse off, or much worse off regarding:
H. Acquiring the skills needed to get a good job.
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} better
\item \verb{[3]} about the same
\item \verb{[4]} worse
\item \verb{[5]} much worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5115/vshow}{GSS Data Explorer page for \code{chldskls}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much better \tab better \tab about the same \tab worse \tab much worse \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 99 \tab 547 \tab 396 \tab 248 \tab 35 \tab 33 \tab 14 \tab 1372 \cr
   Total \tab 0 \tab 99 \tab 547 \tab 396 \tab 248 \tab 35 \tab 33 \tab 14 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{sppregnt}}
}
\concept{Children}
\keyword{variable}
