% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_558.R
\name{chldsup}
\alias{chldsup}
\title{Supervision improving?}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldsup
}
\details{
Question Compared with 10 years ago, would you say that American children today are much better off, better off, about the same, worse off, or much worse off regarding:
E. The supervision and discipline they receive from their parents.
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} better
\item \verb{[3]} about the same
\item \verb{[4]} worse
\item \verb{[5]} much worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5112/vshow}{GSS Data Explorer page for \code{chldsup}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much better \tab better \tab about the same \tab worse \tab much worse \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 21 \tab 90 \tab 457 \tab 668 \tab 94 \tab 29 \tab 13 \tab 1372 \cr
   Total \tab 0 \tab 21 \tab 90 \tab 457 \tab 668 \tab 94 \tab 29 \tab 13 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldeduc}},
\code{\link{chldlove}},
\code{\link{chldmorl}},
\code{\link{chldsafe}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolepa}},
\code{\link{roletchr}}
}
\concept{Children}
\keyword{variable}
