% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chldvig}
\alias{chldvig}
\title{Vignette asked}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldvig
}
\details{
Question Child mental health vignette version number
}
\section{Values}{

\itemize{
\item \verb{[1]} 1:  adhd, john, white, 8
\item \verb{[2]} 2:  adhd, john, white, 14
\item \verb{[3]} 3:  adhd, mary, white, 8
\item \verb{[4]} 4:  adhd, mary, white, 14
\item \verb{[5]} 5:  adhd, john, black, 8
\item \verb{[6]} 6:  adhd, john, black, 14
\item \verb{[7]} 7:  adhd, mary, black, 8
\item \verb{[8]} 8:  adhd, mary, black, 14
\item \verb{[9]} 9:  depression, john, white, 8
\item \verb{[10]} 10: depression, john, white, 14
\item \verb{[11]} 11: depression, mary, white, 8
\item \verb{[12]} 12: depression, mary, white, 14
\item \verb{[13]} 13: depression, john, black, 8
\item \verb{[14]} 14: depression, john, black, 14
\item \verb{[15]} 15: depression, mary, black, 8
\item \verb{[16]} 16: depression, mary, black, 14
\item \verb{[17]} 17: normal, john, white, 8
\item \verb{[18]} 18: normal, john, white, 14
\item \verb{[19]} 19: normal, mary, white, 8
\item \verb{[20]} 20: normal, mary, white, 14
\item \verb{[21]} 21: normal, john, black, 8
\item \verb{[22]} 22: normal, john, black, 14
\item \verb{[23]} 23: normal, mary, black, 8
\item \verb{[24]} 24: normal, mary, black, 14
\item \verb{[25]} 25: asthma, john, white, 8
\item \verb{[26]} 26: asthma, john, white, 14
\item \verb{[27]} 27: asthma, mary, white, 8
\item \verb{[28]} 28: asthma, mary, white, 14
\item \verb{[29]} 29: asthma, john, black, 8
\item \verb{[30]} 30: asthma, john, black, 14
\item \verb{[31]} 31: asthma, mary, black, 8
\item \verb{[32]} 32: asthma, mary, black, 14
\item \verb{[33]} 33: classic depression, john, white, 8
\item \verb{[34]} 34: classic depression, john, white, 14
\item \verb{[35]} 35: classic depression, mary, white, 8
\item \verb{[36]} 36: classic depression, mary, white, 14
\item \verb{[37]} 37: classic depression, john, black, 8
\item \verb{[38]} 38: classic depression, john, black, 14
\item \verb{[39]} 39: classic depression, mary, black, 8
\item \verb{[40]} 40: classic depression, mary, black, 14
\item \verb{[41]} 41: impulsivity, john, white, 8
\item \verb{[42]} 42: impulsivity, john, white, 14
\item \verb{[43]} 43: impulsivity, mary, white, 8
\item \verb{[44]} 44: impulsivity, mary, white, 14
\item \verb{[45]} 45: impulsivity, john, black, 8
\item \verb{[46]} 46: impulsivity, john, black, 14
\item \verb{[47]} 47: impulsivity, mary, black, 8
\item \verb{[48]} 48: impulsivity, mary, black, 14
\item \verb{[49]} 49: perfectionism, john, white, 8
\item \verb{[50]} 50: perfectionism, john, white, 14
\item \verb{[51]} 51: perfectionism, mary, white, 8
\item \verb{[52]} 52: perfectionism, mary, white, 14
\item \verb{[53]} 53: perfectionism, john, black, 8
\item \verb{[54]} 54: perfectionism, john, black, 14
\item \verb{[55]} 55: perfectionism, mary, black, 8
\item \verb{[56]} 56: perfectionism, mary, black, 14
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2917/vshow}{GSS Data Explorer page for \code{chldvig}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1:  adhd, john, white, 8 \tab 2:  adhd, john, white, 14 \tab 3:  adhd, mary, white, 8 \tab 4:  adhd, mary, white, 14 \tab 5:  adhd, john, black, 8 \tab 6:  adhd, john, black, 14 \tab 7:  adhd, mary, black, 8 \tab 8:  adhd, mary, black, 14 \tab 9:  depression, john, white, 8 \tab 10: depression, john, white, 14 \tab 11: depression, mary, white, 8 \tab 12: depression, mary, white, 14 \tab 13: depression, john, black, 8 \tab 14: depression, john, black, 14 \tab 15: depression, mary, black, 8 \tab 16: depression, mary, black, 14 \tab 17: normal, john, white, 8 \tab 18: normal, john, white, 14 \tab 19: normal, mary, white, 8 \tab 20: normal, mary, white, 14 \tab 21: normal, john, black, 8 \tab 22: normal, john, black, 14 \tab 23: normal, mary, black, 8 \tab 24: normal, mary, black, 14 \tab 25: asthma, john, white, 8 \tab 26: asthma, john, white, 14 \tab 27: asthma, mary, white, 8 \tab 28: asthma, mary, white, 14 \tab 29: asthma, john, black, 8 \tab 30: asthma, john, black, 14 \tab 31: asthma, mary, black, 8 \tab 32: asthma, mary, black, 14 \tab not available in this year \tab 33: classic depression, john, white, 8 \tab 34: classic depression, john, white, 14 \tab 35: classic depression, mary, white, 8 \tab 36: classic depression, mary, white, 14 \tab 37: classic depression, john, black, 8 \tab 38: classic depression, john, black, 14 \tab 39: classic depression, mary, black, 8 \tab 40: classic depression, mary, black, 14 \tab 41: impulsivity, john, white, 8 \tab 42: impulsivity, john, white, 14 \tab 43: impulsivity, mary, white, 8 \tab 44: impulsivity, mary, white, 14 \tab 45: impulsivity, john, black, 8 \tab 46: impulsivity, john, black, 14 \tab 47: impulsivity, mary, black, 8 \tab 48: impulsivity, mary, black, 14 \tab 49: perfectionism, john, white, 8 \tab 50: perfectionism, john, white, 14 \tab 51: perfectionism, mary, white, 8 \tab 52: perfectionism, mary, white, 14 \tab 53: perfectionism, john, black, 8 \tab 54: perfectionism, john, black, 14 \tab 55: perfectionism, mary, black, 8 \tab 56: perfectionism, mary, black, 14 \tab Total \cr
   2002 \tab 1372 \tab 36 \tab 45 \tab 39 \tab 47 \tab 42 \tab 45 \tab 47 \tab 39 \tab 42 \tab 40 \tab 50 \tab 43 \tab 54 \tab 48 \tab 48 \tab 49 \tab 40 \tab 49 \tab 48 \tab 42 \tab 48 \tab 44 \tab 38 \tab 43 \tab 47 \tab 44 \tab 34 \tab 41 \tab 36 \tab 38 \tab 46 \tab 41 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2024 \tab 1639 \tab 31 \tab 33 \tab 29 \tab 29 \tab 26 \tab 39 \tab 37 \tab 37 \tab 39 \tab 24 \tab 40 \tab 38 \tab 38 \tab 40 \tab 38 \tab 28 \tab 29 \tab 29 \tab 42 \tab 33 \tab 42 \tab 35 \tab 27 \tab 42 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 31 \tab 29 \tab 41 \tab 34 \tab 35 \tab 31 \tab 37 \tab 40 \tab 32 \tab 34 \tab 33 \tab 43 \tab 25 \tab 31 \tab 44 \tab 26 \tab 44 \tab 34 \tab 37 \tab 32 \tab 42 \tab 42 \tab 33 \tab 35 \tab 3309 \cr
   Total \tab 3011 \tab 67 \tab 78 \tab 68 \tab 76 \tab 68 \tab 84 \tab 84 \tab 76 \tab 81 \tab 64 \tab 90 \tab 81 \tab 92 \tab 88 \tab 86 \tab 77 \tab 69 \tab 78 \tab 90 \tab 75 \tab 90 \tab 79 \tab 65 \tab 85 \tab 47 \tab 44 \tab 34 \tab 41 \tab 36 \tab 38 \tab 46 \tab 41 \tab 0 \tab 31 \tab 29 \tab 41 \tab 34 \tab 35 \tab 31 \tab 37 \tab 40 \tab 32 \tab 34 \tab 33 \tab 43 \tab 25 \tab 31 \tab 44 \tab 26 \tab 44 \tab 34 \tab 37 \tab 32 \tab 42 \tab 42 \tab 33 \tab 35 \tab 6074 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{allergic}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{chembal}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldprob}},
\code{\link{chldsch}},
\code{\link{comknows}},
\code{\link{discipln}},
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{failure}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{illphys}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{medaddct}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mntlill}},
\code{\link{mntlrel}},
\code{\link{nextdoor}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{pryntfam}},
\code{\link{putsoff}},
\code{\link{raised}},
\code{\link{solveown}},
\code{\link{spendeve}},
\code{\link{stressfl}},
\code{\link{sufadult}},
\code{\link{trbllaw}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{zombies}}

Other Mental Health: 
\code{\link{anydoc}},
\code{\link{avoidmh}},
\code{\link{breakdwn}},
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{brkddo1}},
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{brkdtime}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{charactr}},
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_oth}},
\code{\link{chldshop_wht}},
\code{\link{cutself}},
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{depress}},
\code{\link{diagnosd}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{doclist}},
\code{\link{docsui}},
\code{\link{dofirst}},
\code{\link{evbrkdwn}},
\code{\link{evmhp}},
\code{\link{fammhneg}},
\code{\link{fekids5}},
\code{\link{fighpeer}},
\code{\link{foodalc}},
\code{\link{genetics}},
\code{\link{gesttalk}},
\code{\link{godswill}},
\code{\link{govmentl}},
\code{\link{hlthdays}},
\code{\link{hlthinsr}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hospice}},
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imbalnce}},
\code{\link{imprvown}},
\code{\link{imprvown1}},
\code{\link{imprvtrt}},
\code{\link{instrctn}},
\code{\link{knwmhosp}},
\code{\link{knwnproz}},
\code{\link{knwpatnt}},
\code{\link{medaddct}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{meddoc}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentlill}},
\code{\link{mentloth}},
\code{\link{mhaccptd}},
\code{\link{mhbrain}},
\code{\link{mhcause}},
\code{\link{mhchrctr}},
\code{\link{mhclsoth}},
\code{\link{mhcreatv}},
\code{\link{mhdecslf}},
\code{\link{mhdoactv}},
\code{\link{mhdodiet}},
\code{\link{mhdodoc}},
\code{\link{mhdoexrc}},
\code{\link{mhdofam}},
\code{\link{mhdofrgt}},
\code{\link{mhdofrnd}},
\code{\link{mhdogrp}},
\code{\link{mhdoheal}},
\code{\link{mhdoherb}},
\code{\link{mhdohosp}},
\code{\link{mhdolife}},
\code{\link{mhdomhp}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdootc}},
\code{\link{mhdooth}},
\code{\link{mhdopray}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhdorx}},
\code{\link{mhexpoth}},
\code{\link{mhfamshm}},
\code{\link{mhfeelem}},
\code{\link{mhfriend}},
\code{\link{mhfright}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhgvtjob}},
\code{\link{mhhired}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlpoth}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhhrdtlk}},
\code{\link{mhhsptrt}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhkdcare}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhluck}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhmoney}},
\code{\link{mhneihbr}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhnotch}},
\code{\link{mhnotell}},
\code{\link{mhnotpol}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhoutsdr}},
\code{\link{mhphyscl}},
\code{\link{mhprdctv}},
\code{\link{mhproblm}},
\code{\link{mhraised}},
\code{\link{mhreloth}},
\code{\link{mhresoth}},
\code{\link{mhsecret}},
\code{\link{mhseedoc}},
\code{\link{mhseepub}},
\code{\link{mhseroth}},
\code{\link{mhslfshm}},
\code{\link{mhsocial}},
\code{\link{mhstress}},
\code{\link{mhsupwrk}},
\code{\link{mhsymp}},
\code{\link{mhtrtot2}},
\code{\link{mhtrtoth}},
\code{\link{mhtrtslf}},
\code{\link{mhtrusty}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}},
\code{\link{mhupdown}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}},
\code{\link{misswork}},
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}},
\code{\link{mnilwhy3}},
\code{\link{mntlas1}},
\code{\link{mntlas2}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{mntlhlth}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlrel}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}},
\code{\link{morextrm}},
\code{\link{mustdoc}},
\code{\link{mustdoc1}},
\code{\link{musthosp}},
\code{\link{musthosp1}},
\code{\link{mustmed}},
\code{\link{mustmed1}},
\code{\link{negself}},
\code{\link{notreat}},
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}},
\code{\link{otcmed}},
\code{\link{othmhneg}},
\code{\link{painmed}},
\code{\link{physill}},
\code{\link{physill1}},
\code{\link{probfix1}},
\code{\link{probfix2}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozac}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozkid3}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}},
\code{\link{pryntfam}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psyckid3}},
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{psycpsyc}},
\code{\link{putsoff}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}},
\code{\link{rightref}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{rxmed}},
\code{\link{schshoot}},
\code{\link{seemhpub}},
\code{\link{seenmntl}},
\code{\link{selfhelp}},
\code{\link{seriousp}},
\code{\link{smedia}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sphealer}},
\code{\link{spmentl}},
\code{\link{stresses}},
\code{\link{swithlth}},
\code{\link{takelife}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}},
\code{\link{termvig}},
\code{\link{tlkclrgy}},
\code{\link{tlkfam}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}},
\code{\link{upsdowns}},
\code{\link{usedmntl}},
\code{\link{usedproz}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{viglab06}},
\code{\link{viglabel}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigvermy}},
\code{\link{vigversn}},
\code{\link{vigwork}},
\code{\link{wayraise}}

Other Children: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adoption}},
\code{\link{adoptkid}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{aidkids}},
\code{\link{allergic}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{cathsch}},
\code{\link{chembal}},
\code{\link{childs}},
\code{\link{chldben}},
\code{\link{chldcare}},
\code{\link{chldeduc}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldhrt}},
\code{\link{chldidel}},
\code{\link{chldidel_next}},
\code{\link{chldlove}},
\code{\link{chldmore}},
\code{\link{chldmorl}},
\code{\link{chldnum}},
\code{\link{chldpref_next}},
\code{\link{chldprob}},
\code{\link{chldsafe}},
\code{\link{chldsch}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldskls}},
\code{\link{chldsoon}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{chrissch}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{comknows}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{discipln}},
\code{\link{eldersup}},
\code{\link{evkid}},
\code{\link{failure}},
\code{\link{famper4}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{fourkids}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{helpoth}},
\code{\link{hlth8}},
\code{\link{homesch}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{illphys}},
\code{\link{impkids}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpicked}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kid5up}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidhlth}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidlived}},
\code{\link{kidneedy}},
\code{\link{kidnofre}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidnum}},
\code{\link{kidout}},
\code{\link{kidpars}},
\code{\link{kidresp}},
\code{\link{kidskips}},
\code{\link{kidsocst}},
\code{\link{kidssol}},
\code{\link{kidtrble}},
\code{\link{kidxfilm}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{law7}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{mapaid}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{mntlill}},
\code{\link{morextrm}},
\code{\link{mykids}},
\code{\link{nextdoor}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obeythnk}},
\code{\link{obrespct}},
\code{\link{onekid}},
\code{\link{oth3}},
\code{\link{othkids}},
\code{\link{othsch}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{popular}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{raised}},
\code{\link{relsch}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{sellbaby}},
\code{\link{spanking}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spendeve}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}},
\code{\link{spwrkpar}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{stressfl}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{sufadult}},
\code{\link{thnkself}},
\code{\link{threkids}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{trbllaw}},
\code{\link{twokids}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{workhard}},
\code{\link{zombies}}
}
\concept{Children}
\concept{Mental Health}
\concept{Mental Health II}
\keyword{variable}
