% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{chnang2}
\alias{chnang2}
\title{Had a drink or took a pill}
\source{
General Social Survey https://gss.norc.org
}
\description{
chnang2
}
\details{
Question I'm going to show you a list of things that people sometimes do to change their feelings. Did you do any of these things after you got angry or annoyed,... (INTERVIEWER CODE; CONTINUE WITH "Did you do any of these other things in addition to that?" UNTIL RESPONDENT SAYS NO. CIRCLE ALL TECHNIQUES MENTIONED. DO NOT READ ITEMS ON LIST.)
B. Had a drink or took a pill
}
\section{Values}{

\itemize{
\item \verb{[1]} did
\item \verb{[2]} didnt
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1768/vshow}{GSS Data Explorer page for \code{chnang2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab did \tab didnt \tab no answer \tab Total \cr
   1996 \tab 1779 \tab 68 \tab 1046 \tab 11 \tab 2904 \cr
   Total \tab 1779 \tab 68 \tab 1046 \tab 11 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{howangry}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{thnkangr}}
}
\concept{Emotions}
\keyword{variable}
