% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{chngeoth}
\alias{chngeoth}
\title{How often r try to persuade other to share views}
\source{
General Social Survey https://gss.norc.org
}
\description{
chngeoth
}
\details{
Question When you hold a strong opinion about politics, how often do you try to persuade your friends, relatives or fellow workers to share your views?
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4948/vshow}{GSS Data Explorer page for \code{chngeoth}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 180 \tab 427 \tab 471 \tab 386 \tab 3 \tab 5 \tab 2812 \cr
   2014 \tab 1274 \tab 70 \tab 339 \tab 386 \tab 441 \tab 24 \tab 4 \tab 2538 \cr
   Total \tab 2614 \tab 250 \tab 766 \tab 857 \tab 827 \tab 27 \tab 9 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
