% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{chngtime}
\alias{chngtime}
\title{How often change start and end times}
\source{
General Social Survey https://gss.norc.org
}
\description{
chngtime
}
\details{
Question How often are you allowed to change your starting and quitting times on a daily basis?
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7854/vshow}{GSS Data Explorer page for \code{chngtime}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1540 \tab 675 \tab 413 \tab 323 \tab 549 \tab 12 \tab 15 \tab 17 \tab 3544 \cr
   Total \tab 1540 \tab 675 \tab 413 \tab 323 \tab 549 \tab 12 \tab 15 \tab 17 \tab 3544 \cr
}
}

\keyword{variable}
