% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{chores1}
\alias{chores1}
\title{Help for household tasks #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
chores1
}
\details{
Question First, there are some household and garden jobs you can''t really Do alone --for example, you may need someone to hold a ladder, or to help you move furniture.
A. Who would you turn to first for help?
B. Who would you turn to second?
Please CIRCLE ONLY ONE FIRST CHOICE NUMBER AND ONE SECOND CHOICE NUMBER.
A. First Choice
}
\section{Values}{

\itemize{
\item \code{1} spouse
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other relative
\item \code{9} closest friend
\item \code{10} other friend
\item \code{11} neighbor
\item \code{12} coworker
\item \code{13} social service
\item \code{14} paid help
\item \code{15} other
\item \code{16} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4169/vshow}{GSS Data Explorer page for \code{chores1}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab neighbor \tab coworker \tab social service \tab paid help \tab other \tab no one \tab no answer \tab Total \cr
   1986 \tab 42 \tab 701 \tab 35 \tab 51 \tab 61 \tab 152 \tab 25 \tab 69 \tab 49 \tab 123 \tab 17 \tab 68 \tab 8 \tab 3 \tab 25 \tab 4 \tab 18 \tab 19 \tab 1470 \cr
   Total \tab 42 \tab 701 \tab 35 \tab 51 \tab 61 \tab 152 \tab 25 \tab 69 \tab 49 \tab 123 \tab 17 \tab 68 \tab 8 \tab 3 \tab 25 \tab 4 \tab 18 \tab 19 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
