% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{chosedoc}
\alias{chosedoc}
\title{R had enough choices among docs}
\source{
General Social Survey https://gss.norc.org
}
\description{
chosedoc
}
\details{
Question Do you feel that you had enough choices among different doctors in your health insurance plan when you chose your current doctor?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2744/vshow}{GSS Data Explorer page for \code{chosedoc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 375 \tab 1927 \tab 428 \tab 26 \tab 9 \tab 2765 \cr
   Total \tab 375 \tab 1927 \tab 428 \tab 26 \tab 9 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
