% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{chrissch}
\alias{chrissch}
\title{Children attend christian school}
\source{
General Social Survey https://gss.norc.org
}
\description{
chrissch
}
\details{
Question Do/did all of your children attend public school for all of their elementary and high school education?
A. Which types of school do/did your children attend?
3. Christian school (not Catholic)
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1150/vshow}{GSS Data Explorer page for \code{chrissch}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 2634 \tab 36 \tab 119 \tab 43 \tab 2832 \cr
   2000 \tab 2540 \tab 50 \tab 199 \tab 28 \tab 2817 \cr
   Total \tab 5174 \tab 86 \tab 318 \tab 71 \tab 5649 \cr
}
}

\seealso{
Other Religion: 
\code{\link{denomsch}},
\code{\link{implives}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}}
}
\concept{Religion}
\keyword{variable}
