% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_633.R
\name{chrstsav}
\alias{chrstsav}
\title{Only those who accept jesus as savior can achieve salvation}
\source{
General Social Survey https://gss.norc.org
}
\description{
chrstsav
}
\details{
Question Only those who accept Jesus Christ alone as their Savior can achieve eternal salvation.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7533/vshow}{GSS Data Explorer page for \code{chrstsav}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 2258 \tab 565 \tab 307 \tab 317 \tab 150 \tab 145 \tab 4 \tab 163 \tab 25 \tab 98 \tab 4032 \cr
   Total \tab 2258 \tab 565 \tab 307 \tab 317 \tab 150 \tab 145 \tab 4 \tab 163 \tab 25 \tab 98 \tab 4032 \cr
}
}

\keyword{variable}
