% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{churchtx}
\alias{churchtx}
\title{Membership in church-related groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
churchtx
}
\details{
Question Besides belonging to a congregation or religion, do you belong to any church-related groups such as a fellowship or study groups, a choir, board, or committee?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3108/vshow}{GSS Data Explorer page for \code{churchtx}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2352 \tab 239 \tab 215 \tab 1 \tab 5 \tab 2812 \cr
   Total \tab 2352 \tab 239 \tab 215 \tab 1 \tab 5 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{infrmgrp}},
\code{\link{typfrat}},
\code{\link{typpolit}},
\code{\link{typserv}},
\code{\link{typvet}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
