% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{churhgrp}
\alias{churhgrp}
\title{Is church group church itself}
\source{
General Social Survey https://gss.norc.org
}
\description{
churhgrp
}
\details{
Question You said you were a member of a church affiliated group. Is that group or organization the church (synagogue) itself, or some other group related to the church?
}
\section{Values}{

\itemize{
\item \verb{[1]} church
\item \verb{[2]} related group
\item \verb{[3]} both
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1012/vshow}{GSS Data Explorer page for \code{churhgrp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab church \tab related group \tab both \tab no answer \tab Total \cr
   1987 \tab 1227 \tab 309 \tab 181 \tab 77 \tab 25 \tab 1819 \cr
   Total \tab 1227 \tab 309 \tab 181 \tab 77 \tab 25 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{caregrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{mostact}},
\code{\link{servegrp}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
