% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_116.R
\name{churhmem}
\alias{churhmem}
\title{Is r a member of church or synagogue}
\source{
General Social Survey https://gss.norc.org
}
\description{
churhmem
}
\details{
Question Are you, yourself, a member of a church or synagogue?
Yes........(ASK CHURHACT)…....1
No..........(GO TO  TVRELIG)........2
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1073/vshow}{GSS Data Explorer page for \code{churhmem}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   1988 \tab - \tab 904 \tab 577 \tab 1481 \cr
   Total \tab 0 \tab 904 \tab 577 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhact}},
\code{\link{god}},
\code{\link{grace16}},
\code{\link{join16}},
\code{\link{lapsed}},
\code{\link{reborn}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{tvrelig}}
}
\concept{Religion}
\keyword{variable}
