% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{churhpow}
\alias{churhpow}
\title{Do churches have too much power?}
\source{
General Social Survey https://gss.norc.org
}
\description{
churhpow
}
\details{
Question Do you think that churches and religious organizations in this country have too much power or too little power?
}
\section{Values}{

\itemize{
\item \verb{[1]} far too much power
\item \verb{[2]} too much power
\item \verb{[3]} about the right amount of power
\item \verb{[4]} too little power
\item \verb{[5]} far too little power
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4649/vshow}{GSS Data Explorer page for \code{churhpow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab far too much power \tab too much power \tab about the right amount of power \tab too little power \tab far too little power \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 88 \tab 185 \tab 698 \tab 178 \tab 36 \tab 141 \tab 33 \tab 1517 \cr
   1998 \tab 1548 \tab 68 \tab 186 \tab 596 \tab 182 \tab 50 \tab 157 \tab 45 \tab 2832 \cr
   2008 \tab 658 \tab 88 \tab 336 \tab 649 \tab 201 \tab 23 \tab 58 \tab 10 \tab 2023 \cr
   2018 \tab 1173 \tab 83 \tab 320 \tab 481 \tab 216 \tab 16 \tab 46 \tab 13 \tab 2348 \cr
   Total \tab 3537 \tab 327 \tab 1027 \tab 2424 \tab 777 \tab 125 \tab 402 \tab 101 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
