% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cideknew}
\alias{cideknew}
\title{#People r knows murdered in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
cideknew
}
\details{
Question Within the past 12 months, how many people have you known personally that were victims of homicide.
}
\section{Values}{

\itemize{
\item \verb{[7]} seven or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4990/vshow}{GSS Data Explorer page for \code{cideknew}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab seven or more \tab no answer \tab Total \cr
   1988 \tab - \tab 1325 \tab 95 \tab 36 \tab 13 \tab 4 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1481 \cr
   1989 \tab - \tab 1375 \tab 103 \tab 34 \tab 16 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab 1537 \cr
   1990 \tab - \tab 1244 \tab 81 \tab 27 \tab 11 \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab 1372 \cr
   Total \tab 0 \tab 3944 \tab 279 \tab 97 \tab 40 \tab 12 \tab 7 \tab 2 \tab 2 \tab 7 \tab 4390 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
