% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{ciderace}
\alias{ciderace}
\title{Race of murder victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
ciderace
}
\details{
Question We would like to know a few things about that person.
C. What was that person''s race?  Was it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \verb{[1]} black
\item \verb{[2]} white
\item \verb{[3]} hispanic
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4994/vshow}{GSS Data Explorer page for \code{ciderace}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab black \tab white \tab hispanic \tab other \tab no answer \tab not available in this year \tab Total \cr
   1988 \tab 1325 \tab 48 \tab 81 \tab 18 \tab 7 \tab 2 \tab - \tab 1481 \cr
   1989 \tab 1375 \tab 45 \tab 100 \tab 14 \tab - \tab 3 \tab - \tab 1537 \cr
   1990 \tab 1244 \tab 51 \tab 62 \tab 8 \tab 5 \tab 2 \tab - \tab 1372 \cr
   Total \tab 3944 \tab 144 \tab 243 \tab 40 \tab 12 \tab 7 \tab 0 \tab 4390 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssch}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}},
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suiage3}},
\code{\link{suiknew}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Homicide: 
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}}
}
\concept{Homicide}
\concept{Victims & AIDS}
\keyword{variable}
