% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cidesex}
\alias{cidesex}
\title{Sex of murder victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidesex
}
\details{
Question We would like to know a few things about that person.
A. Was that person male or female?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4992/vshow}{GSS Data Explorer page for \code{cidesex}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1988 \tab 1325 \tab 121 \tab 31 \tab 4 \tab 1481 \cr
   1989 \tab 1375 \tab 116 \tab 43 \tab 3 \tab 1537 \cr
   1990 \tab 1244 \tab 94 \tab 32 \tab 2 \tab 1372 \cr
   Total \tab 3944 \tab 331 \tab 106 \tab 9 \tab 4390 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
