% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{cidesex3}
\alias{cidesex3}
\title{Sex of murder victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidesex3
}
\details{
Question We would like to know a few things about that person.
A. Was that person male or female?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5002/vshow}{GSS Data Explorer page for \code{cidesex3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab not available in this year \tab Total \cr
   1989 \tab 1512 \tab 19 \tab 3 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 1352 \tab 13 \tab 5 \tab 2 \tab - \tab 1372 \cr
   Total \tab 2864 \tab 32 \tab 8 \tab 5 \tab 0 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssch}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}},
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suiage3}},
\code{\link{suiknew}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Homicide: 
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}}
}
\concept{Homicide}
\concept{Victims & AIDS}
\keyword{variable}
