% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{cidewho3}
\alias{cidewho3}
\title{R's relationship to 3rd closest murderee}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidewho3
}
\details{
Question Think about the person you knew next best who was a victim of homicide. Please tell me the letter of the category On the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \verb{[1]} husband or wife
\item \verb{[2]} partner or lover
\item \verb{[3]} son or daughter
\item \verb{[4]} other relative
\item \verb{[5]} friend
\item \verb{[6]} neighbor
\item \verb{[7]} co-worker
\item \verb{[8]} acquaintance
\item \verb{[9]} patient
\item \verb{[10]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5001/vshow}{GSS Data Explorer page for \code{cidewho3}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab other relative \tab friend \tab co-worker \tab acquaintance \tab patient \tab no answer \tab neighbor \tab not available in this year \tab Total \cr
   1989 \tab 1512 \tab 1 \tab 9 \tab 1 \tab 10 \tab 1 \tab 3 \tab - \tab - \tab 1537 \cr
   1990 \tab 1352 \tab 2 \tab 6 \tab 2 \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 1372 \cr
   Total \tab 2864 \tab 3 \tab 15 \tab 3 \tab 16 \tab 1 \tab 5 \tab 2 \tab 0 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssch}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suiage3}},
\code{\link{suiknew}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Homicide: 
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Homicide}
\concept{Victims & AIDS}
\keyword{variable}
