% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{cigweek}
\alias{cigweek}
\title{Has r smoked cigarettes in last week}
\source{
General Social Survey https://gss.norc.org
}
\description{
cigweek
}
\details{
Question Do you smoke?
C. Have you, yourself, smoked any cigarettes in the past week?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/535/vshow}{GSS Data Explorer page for \code{cigweek}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1977 \tab 932 \tab 571 \tab 6 \tab 21 \tab 1530 \cr
   Total \tab 932 \tab 571 \tab 6 \tab 21 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}
}
\concept{Core}
\keyword{variable}
