% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{citworld}
\alias{citworld}
\title{I am a citizen of the world}
\source{
General Social Survey https://gss.norc.org
}
\description{
citworld
}
\details{
Question How much do you agree or disagree with the following statements? (Please, select one response)
E. I feel more like a citizen of the world than of any country.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4887/vshow}{GSS Data Explorer page for \code{citworld}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1264 \tab 68 \tab 287 \tab 222 \tab 529 \tab 129 \tab 36 \tab 3 \tab 2538 \cr
   Total \tab 1264 \tab 68 \tab 287 \tab 222 \tab 529 \tab 129 \tab 36 \tab 3 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
