% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{civic}
\alias{civic}
\title{Does r follow public affairs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
civic
}
\details{
Question Some people seem to follow what's going on in government and public affairs most of the time, whether there's an election going on or not. Others aren't that interested. Would you say you follow what's going on in government and public affairs most of the time, some of the time, only now and then, or hardly at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} most of the time
\item \verb{[2]} some of the time
\item \verb{[3]} now and then
\item \verb{[4]} hardly at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/773/vshow}{GSS Data Explorer page for \code{civic}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most of the time \tab some of the time \tab now and then \tab hardly at all \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 632 \tab 649 \tab 366 \tab 192 \tab 11 \tab 10 \tab 1860 \cr
   Total \tab 0 \tab 632 \tab 649 \tab 366 \tab 192 \tab 11 \tab 10 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\keyword{variable}
