% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{class1}
\alias{class1}
\title{Subjective class identification (2021)}
\source{
General Social Survey https://gss.norc.org
}
\description{
class1
}
\details{
Question Most people see themselves as belonging to a particular class. Please tell me which social class you would say you belong to?
}
\section{Values}{

\itemize{
\item \code{1} lower class
\item \code{2} working class
\item \code{3} lower middle class
\item \code{4} middle class
\item \code{5} upper middle class
\item \code{6} upper class
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7520/vshow}{GSS Data Explorer page for \code{class1}}.
}

\keyword{variable}
