% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{classics}
\alias{classics}
\title{Hs and coll studts read too many classics}
\source{
General Social Survey https://gss.norc.org
}
\description{
classics
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
C. High schools and colleges make students spend too much time reading "classics" that have little relevance in today's world
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1424/vshow}{GSS Data Explorer page for \code{classics}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 59 \tab 483 \tab 761 \tab 128 \tab 160 \tab 15 \tab 1606 \cr
   Total \tab 0 \tab 59 \tab 483 \tab 761 \tab 128 \tab 160 \tab 15 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{judgeart}},
\code{\link{modpaint}},
\code{\link{pclit}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}}
}
\concept{Culture}
\keyword{variable}
