% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{classy}
\alias{classy}
\title{Subjective social class-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
classy
}
\details{
Question B. If you were asked to use one of five names for your social class, which would you say you belong in: the lower class, the working class, the middle class, the upper middle class, or the upper class?
}
\section{Values}{

\itemize{
\item \verb{[1]} lower class
\item \verb{[2]} working class
\item \verb{[3]} middle class
\item \verb{[4]} upper middle class
\item \verb{[5]} upper class
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/569/vshow}{GSS Data Explorer page for \code{classy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab lower class \tab working class \tab middle class \tab upper middle class \tab upper class \tab no answer \tab Total \cr
   1973 \tab 753 \tab 32 \tab 308 \tab 308 \tab 79 \tab 18 \tab 6 \tab 1504 \cr
   Total \tab 753 \tab 32 \tab 308 \tab 308 \tab 79 \tab 18 \tab 6 \tab 1504 \cr
}
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()},
\code{\link{rank10}}
}
\concept{Core}
\keyword{variable}
