% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_54.R
\name{clean}
\alias{clean}
\title{Neat and clean}
\source{
General Social Survey https://gss.norc.org
}
\description{
clean
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
4. That he is neat and clean
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 most desirable
\item \verb{[2]} 3 most desirable
\item \verb{[3]} not mentioned
\item \verb{[4]} 3 least desirable
\item \verb{[5]} 1 least desirable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1975, 1976, 1978, 1980, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/477/vshow}{GSS Data Explorer page for \code{clean}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 11 \tab 121 \tab 823 \tab 348 \tab 197 \tab 4 \tab - \tab 1504 \cr
   1975 \tab - \tab 12 \tab 97 \tab 820 \tab 369 \tab 182 \tab 10 \tab - \tab 1490 \cr
   1976 \tab - \tab 9 \tab 98 \tab 797 \tab 380 \tab 205 \tab 10 \tab - \tab 1499 \cr
   1978 \tab - \tab 13 \tab 123 \tab 854 \tab 346 \tab 184 \tab 10 \tab 2 \tab 1532 \cr
   1980 \tab 975 \tab 1 \tab 29 \tab 265 \tab 126 \tab 68 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 12 \tab 94 \tab 825 \tab 433 \tab 215 \tab 20 \tab - \tab 1599 \cr
   Total \tab 975 \tab 58 \tab 562 \tab 4384 \tab 2002 \tab 1051 \tab 57 \tab 3 \tab 9092 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicable}},
\code{\link{consider}},
\code{\link{control}},
\code{\link{honest}},
\code{\link{judgment}},
\code{\link{obeys}},
\code{\link{responsi}},
\code{\link{role}},
\code{\link{success}}
}
\concept{Core}
\keyword{variable}
