% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{clergvte}
\alias{clergvte}
\title{Clergy shouldn't try to influence voters}
\source{
General Social Survey https://gss.norc.org
}
\description{
clergvte
}
\details{
Question How much do you agree or disagree with each of the following…
B. Religious leaders should not try to influence how people vote in elections.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4646/vshow}{GSS Data Explorer page for \code{clergvte}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 395 \tab 448 \tab 223 \tab 201 \tab 31 \tab 39 \tab 22 \tab 1517 \cr
   1998 \tab 1548 \tab 447 \tab 367 \tab 210 \tab 150 \tab 55 \tab 21 \tab 34 \tab 2832 \cr
   2008 \tab 658 \tab 607 \tab 379 \tab 186 \tab 146 \tab 37 \tab 8 \tab 2 \tab 2023 \cr
   2018 \tab 1173 \tab 463 \tab 372 \tab 174 \tab 110 \tab 42 \tab 6 \tab 8 \tab 2348 \cr
   Total \tab 3537 \tab 1912 \tab 1566 \tab 793 \tab 607 \tab 165 \tab 74 \tab 66 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
