% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{clmtcaus}
\alias{clmtcaus}
\title{Climate has been changing due to human activity}
\source{
General Social Survey https://gss.norc.org
}
\description{
clmtcaus
}
\details{
Question There has been a lot of discussion about the world's climate and the idea it has been changing in recent decades. Which of the following statements comes closest to your opinion?
}
\section{Values}{

\itemize{
\item \verb{[1]} has not been changing
\item \verb{[2]} has been changing mostly due to natural processes
\item \verb{[3]} has been changing about equally due to natural processes and human activity
\item \verb{[4]} has been changing mostly due to human activity
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7486/vshow}{GSS Data Explorer page for \code{clmtcaus}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab has not been changing \tab has been changing mostly due to natural processes \tab has been changing about equally due to natural processes and human activity \tab has been changing mostly due to human activity \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 35 \tab 240 \tab 639 \tab 856 \tab 90 \tab 160 \tab 9 \tab 4032 \cr
   Total \tab 2003 \tab 35 \tab 240 \tab 639 \tab 856 \tab 90 \tab 160 \tab 9 \tab 4032 \cr
}
}

\keyword{variable}
