% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{cloning}
\alias{cloning}
\title{Cloning makes identical copy}
\source{
General Social Survey https://gss.norc.org
}
\description{
cloning
}
\details{
Question The cloning of living things produces genetically identical copies.
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3499/vshow}{GSS Data Explorer page for \code{cloning}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Genetic Testing}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
