% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_94.R
\name{close15}
\alias{close15}
\title{How close are persons number 1 and 5?}
\source{
General Social Survey https://gss.norc.org
}
\description{
close15
}
\details{
Question IF LESS THAN 5 NAMES MENTIONED CROSS OUT UNUSED BOXES.
Please think about the relations between the people you just mentioned. Some of them may be total strangers in the sense that they wouldn't recognize each other if they bumped into each other on the street. Others may be especially close, as close or closer to each other as they are to you.
First, think about NAME 1 and NAME 2.
ASK  CLOSE1 FOR FIRST PAIR.
A. Are ____ and ____ total strangers?
IF YES.......(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DoWN)
IF NO........(ASK CLOSE12, CLOSE13, CLOSE14, CLOSE15, CLOSE 23, CLOSE24, CLOSE25, CLOSE34, CLOSE35, CLOSE45)
B. Are they especially close?  PROBE:  As close or closer to each other as they are to you?
IF YES.......(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DOWN)
IF NO........(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DOWN)
4. Closeness of First and Fifth Persons
}
\section{Values}{

\itemize{
\item \code{1} especially close
\item \code{2} know each other
\item \code{3} total strangers
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/858/vshow}{GSS Data Explorer page for \code{close15}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab especially close \tab know each other \tab total strangers \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 132 \tab 143 \tab 87 \tab 19 \tab 1534 \cr
   2004 \tab 2600 \tab 57 \tab 65 \tab 44 \tab 46 \tab 2812 \cr
   Total \tab 3753 \tab 189 \tab 208 \tab 131 \tab 65 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close23}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}}
}
\concept{Social Networks}
\keyword{variable}
