% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_95.R
\name{close25}
\alias{close25}
\title{How close are persons number 2 and 5?}
\source{
General Social Survey https://gss.norc.org
}
\description{
close25
}
\details{
Question IF LESS THAN 5 NAMES MENTIONED CROSS OUT UNUSED BOXES.
Please think about the relations between the people you just mentioned. Some of them may be total strangers in the sense that they wouldn't recognize each other if they bumped into each other on the street. Others may be especially close, as close or closer to each other as they are to you.
First, think about NAME 1 and NAME 2.
ASK  CLOSE1 FOR FIRST PAIR.
A. Are ____ and ____ total strangers?
IF YES.......(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DoWN)
IF NO........(ASK CLOSE12, CLOSE13, CLOSE14, CLOSE15, CLOSE 23, CLOSE24, CLOSE25, CLOSE34, CLOSE35, CLOSE45)
B. Are they especially close?  PROBE:  As close or closer to each other as they are to you?
IF YES.......(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DOWN)
IF NO........(ASK NAME1, NAME2, NAME3, NAME4, NAME5 FOR NEXT PAIR DOWN) n7. Closeness of Second and Fifth Persons
}
\section{Values}{

\itemize{
\item \verb{[1]} especially close
\item \verb{[2]} know each other
\item \verb{[3]} total strangers
\item \verb{[7]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/861/vshow}{GSS Data Explorer page for \code{close25}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab especially close \tab know each other \tab total strangers \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 124 \tab 128 \tab 109 \tab 20 \tab 1534 \cr
   2004 \tab 2600 \tab 50 \tab 76 \tab 39 \tab 47 \tab 2812 \cr
   Total \tab 3753 \tab 174 \tab 204 \tab 148 \tab 67 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close24}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}}
}
\concept{Social Networks}
\keyword{variable}
