% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_222.R
\name{closeto1}
\alias{closeto1}
\title{How close r is to person 1}
\source{
General Social Survey https://gss.norc.org
}
\description{
closeto1
}
\details{
Question On a scale from 1 to 10, with 1 being not at all close and 10 being very close, how close are you to (NAME 1), the first person?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - not at all close
\item \verb{[10]} 10 - very close
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7110/vshow}{GSS Data Explorer page for \code{closeto1}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all close \tab 10 - very close \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1751 \tab 80 \tab 143 \tab 43 \tab 46 \tab 37 \tab 68 \tab 37 \tab 46 \tab 59 \tab 33 \tab 2 \tab 3 \tab 2348 \cr
   Total \tab 1751 \tab 80 \tab 143 \tab 43 \tab 46 \tab 37 \tab 68 \tab 37 \tab 46 \tab 59 \tab 33 \tab 2 \tab 3 \tab 2348 \cr
}
}

\keyword{variable}
