% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_222.R
\name{closeto5}
\alias{closeto5}
\title{How close r is to person 5}
\source{
General Social Survey https://gss.norc.org
}
\description{
closeto5
}
\details{
Question On a scale from 1 to 10, with 1 being not at all close and 10 being very close, how close are you to (NAME 5), the fifth person?
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all close
\item \code{10} 10 - very close
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7114/vshow}{GSS Data Explorer page for \code{closeto5}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all close \tab 10 - very close \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2245 \tab 19 \tab 8 \tab 10 \tab 5 \tab 12 \tab 9 \tab 8 \tab 13 \tab 6 \tab 5 \tab 2 \tab 6 \tab 2348 \cr
   Total \tab 2245 \tab 19 \tab 8 \tab 10 \tab 5 \tab 12 \tab 9 \tab 8 \tab 13 \tab 6 \tab 5 \tab 2 \tab 6 \tab 2348 \cr
}
}

\keyword{variable}
