% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_431.R
\name{clrggen1}
\alias{clrggen1}
\title{Sex of leader1}
\source{
General Social Survey https://gss.norc.org
}
\description{
clrggen1
}
\details{
Question A. Was this leader male or female?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3951/vshow}{GSS Data Explorer page for \code{clrggen1}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab male \tab Total \cr
   2008 \tab 2001 \tab 22 \tab 2023 \cr
   Total \tab 2001 \tab 22 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{hushrel1}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
