% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{clrgmode}
\alias{clrgmode}
\title{Was saq self administered or interview administered}
\source{
General Social Survey https://gss.norc.org
}
\description{
clrgmode
}
\details{
Question Was the SAQ respondent administered or Interviewer administered?
}
\section{Values}{

\itemize{
\item \verb{[1]} interviewed
\item \verb{[2]} self-administered
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3968/vshow}{GSS Data Explorer page for \code{clrgmode}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab interviewed \tab self-administered \tab Total \cr
   2008 \tab 233 \tab 531 \tab 1259 \tab 2023 \cr
   Total \tab 233 \tab 531 \tab 1259 \tab 2023 \cr
}
}

\concept{Clergy/Sex}
\keyword{variable}
