% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{clsetown}
\alias{clsetown}
\title{How close do you feel to your town or city}
\source{
General Social Survey https://gss.norc.org
}
\description{
clsetown
}
\details{
Question How close do you feel to…
2. Your town or city
}
\section{Values}{

\itemize{
\item \code{1} very close
\item \code{2} close
\item \code{3} not very close
\item \code{4} not close at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4817/vshow}{GSS Data Explorer page for \code{clsetown}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very close \tab close \tab not very close \tab not close at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 173 \tab 623 \tab 413 \tab 108 \tab 23 \tab 27 \tab 2904 \cr
   2004 \tab 1596 \tab 268 \tab 540 \tab 283 \tab 84 \tab 41 \tab - \tab 2812 \cr
   2014 \tab 1264 \tab 306 \tab 519 \tab 281 \tab 119 \tab 47 \tab 2 \tab 2538 \cr
   Total \tab 4397 \tab 747 \tab 1682 \tab 977 \tab 311 \tab 111 \tab 29 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{clsenei}}
}
\concept{ISSP National Identity}
\keyword{variable}
