% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{clseusa}
\alias{clseusa}
\title{How close do you feel to america}
\source{
General Social Survey https://gss.norc.org
}
\description{
clseusa
}
\details{
Question How close do you feel to…
4. America
}
\section{Values}{

\itemize{
\item \code{1} very close
\item \code{2} close
\item \code{3} not very close
\item \code{4} not close at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4819/vshow}{GSS Data Explorer page for \code{clseusa}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very close \tab close \tab not very close \tab not close at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 463 \tab 596 \tab 200 \tab 48 \tab 31 \tab 29 \tab 2904 \cr
   2004 \tab 1596 \tab 623 \tab 442 \tab 102 \tab 23 \tab 25 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 539 \tab 518 \tab 141 \tab 41 \tab 32 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 1625 \tab 1556 \tab 443 \tab 112 \tab 88 \tab 33 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{onenatn}}
}
\concept{ISSP National Identity}
\keyword{variable}
