% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{clsrltv}
\alias{clsrltv}
\title{Who are close relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
clsrltv
}
\details{
Question People may have different perceptions about who they consider to be close relatives. Imagine a person living with a partner and their children. Which of the following options comes closest to who you would consider to be his/her close relatives?
}
\section{Values}{

\itemize{
\item \verb{[1]} children and partner only
\item \verb{[2]} children, partner, parents and siblings
\item \verb{[3]} children, partner, parents, siblings, cousins, aunts/uncles, nephews/nieces
\item \verb{[4]} children, partner, parents, siblings, cousins, aunts/uncles, nephews/nieces, but others as well
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7777/vshow}{GSS Data Explorer page for \code{clsrltv}}.

Counts by year:\tabular{lllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab Total \cr
   2022 \tab 2587 \tab 97 \tab 369 \tab 282 \tab 209 \tab 3544 \cr
   Total \tab 2587 \tab 97 \tab 369 \tab 282 \tab 209 \tab 3544 \cr
}
}

\keyword{variable}
