% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_287.R
\name{clssmtes}
\alias{clssmtes}
\title{Use clssmtes}
\source{
General Social Survey https://gss.norc.org
}
\description{
clssmtes
}
\details{
Question (Which of the following social networking or social media sites are you a member or regular user of?)
Classmates.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6280/vshow}{GSS Data Explorer page for \code{clssmtes}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 1483 \tab 77 \tab 1295 \tab 12 \tab 2867 \cr
   Total \tab 1483 \tab 77 \tab 1295 \tab 12 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{flickr}},
\code{\link{pinterst}},
\code{\link{snsmnew}},
\code{\link{snsmot2a}},
\code{\link{snsmot2b}},
\code{\link{snsmot2c}},
\code{\link{snsmoth1}},
\code{\link{vine}}
}
\concept{Social Media Usage}
\keyword{variable}
