% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_539.R
\name{cntctgov}
\alias{cntctgov}
\title{Contacted politician or civil servant to express view}
\source{
General Social Survey https://gss.norc.org
}
\description{
cntctgov
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the past year, whether you have done it in the more distant past, whether you have not done it but might do it, or have not done it and would never, under any circumstances, do it.
E. Contacted, or attempted to contact, a politician or a civil servant to express your views
}
\section{Values}{

\itemize{
\item \code{1} have done it in the past yr
\item \code{2} have done it in the more distant past
\item \code{3} have not done it but might do it
\item \code{4} have not done it and would never do it
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4924/vshow}{GSS Data Explorer page for \code{cntctgov}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{polfunds}},
\code{\link{racmeet}},
\code{\link{revmeet}},
\code{\link{signdpet}},
\code{\link{usemedia}}

Other Political: 
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{polfunds}},
\code{\link{signdpet}},
\code{\link{usemedia}}
}
\concept{ISSP Citizenship}
\concept{Political}
\keyword{variable}
