% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{cntctkid}
\alias{cntctkid}
\title{How often r contacts their most-contacted adult child}
\source{
General Social Survey https://gss.norc.org
}
\description{
cntctkid
}
\details{
Question Think about your adult child you have contact with most frequently: How often do you have contact with this child aged at least 18, either face-to-face, by phone, internet or any other communication device?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several times a week
\item \verb{[3]} once a week
\item \verb{[4]} two to three times a month
\item \verb{[5]} once a month
\item \verb{[6]} several times a year
\item \verb{[7]} less often
\item \verb{[8]} never
\item \verb{[9]} i do not have any adult children
\item \verb{[10]} the adult child i have contact with most frequently lives with me
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7267/vshow}{GSS Data Explorer page for \code{cntctkid}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab daily \tab several times a week \tab once a week \tab two to three times a month \tab once a month \tab several times a year \tab less often \tab never \tab i do not have any adult children \tab the adult child i have contact with most frequently lives with me \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 264 \tab 151 \tab 63 \tab 38 \tab 16 \tab 13 \tab 7 \tab 23 \tab 563 \tab 18 \tab 15 \tab 2 \tab 2348 \cr
   Total \tab 1175 \tab 264 \tab 151 \tab 63 \tab 38 \tab 16 \tab 13 \tab 7 \tab 23 \tab 563 \tab 18 \tab 15 \tab 2 \tab 2348 \cr
}
}

\keyword{variable}
