% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{cntctsib}
\alias{cntctsib}
\title{How often r contacts their most-contacted sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
cntctsib
}
\details{
Question Think about the brother or sister you have contact with most frequently: How often do you have contact with that brother or sister, either face-to-face, by phone, internet or any other communication device?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several times a week
\item \verb{[3]} once a week
\item \verb{[4]} two to three times a month
\item \verb{[5]} once a month
\item \verb{[6]} several times a year
\item \verb{[7]} less often
\item \verb{[8]} never
\item \verb{[9]} i do not have brothers or sisters
\item \verb{[10]} the brother or sister i have contact with most frequently lives with me
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7266/vshow}{GSS Data Explorer page for \code{cntctsib}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab daily \tab several times a week \tab once a week \tab two to three times a month \tab once a month \tab several times a year \tab less often \tab never \tab i do not have brothers or sisters \tab the brother or sister i have contact with most frequently lives with me \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 176 \tab 258 \tab 164 \tab 193 \tab 118 \tab 71 \tab 45 \tab 61 \tab 71 \tab 9 \tab 5 \tab 2 \tab 2348 \cr
   Total \tab 1175 \tab 176 \tab 258 \tab 164 \tab 193 \tab 118 \tab 71 \tab 45 \tab 61 \tab 71 \tab 9 \tab 5 \tab 2 \tab 2348 \cr
}
}

\keyword{variable}
