% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{cntrlife}
\alias{cntrlife}
\title{Control r feels over own life}
\source{
General Social Survey https://gss.norc.org
}
\description{
cntrlife
}
\details{
Question Some people feel they have completely free choice and control over the way their lives turn out, and other people feel that what they themselves do has no real effect on what happens to them. Please use this scale to indicate how much freedom of choice and control you feel you have over the way your life turns out.
}
\section{Values}{

\itemize{
\item \verb{[1]} no choice and control
\item \verb{[4]} some choice and control
\item \verb{[7]} a great deal of choice and control
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2676/vshow}{GSS Data Explorer page for \code{cntrlife}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab no choice and control \tab 2 \tab 3 \tab some choice and control \tab 5 \tab 6 \tab a great deal of choice and control \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 9 \tab 19 \tab 35 \tab 289 \tab 201 \tab 384 \tab 448 \tab 19 \tab 15 \tab 2817 \cr
   Total \tab 1398 \tab 9 \tab 19 \tab 35 \tab 289 \tab 201 \tab 384 \tab 448 \tab 19 \tab 15 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{expunpop}},
\code{\link{freeprss}},
\code{\link{wlthpov}}
}
\concept{Freedom}
\keyword{variable}
