% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{cobarate}
\alias{cobarate}
\title{Rate of ba degree earning at college partner attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
cobarate
}
\details{
Question In what year did your partner receive the (COLLEGE NAME) degree?
K. Graduation Rate
}
\section{Values}{

\itemize{
\item \verb{[1]} 0-39.9pct
\item \verb{[2]} 40-57.9pct
\item \verb{[3]} 58pct or over
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6164/vshow}{GSS Data Explorer page for \code{cobarate}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 0-39.9pct \tab 40-57.9pct \tab 58pct or over \tab no answer \tab Total \cr
   2012 \tab 1940 \tab 6 \tab 8 \tab 20 \tab - \tab 1974 \cr
   2014 \tab 2515 \tab 2 \tab 6 \tab 15 \tab - \tab 2538 \cr
   2016 \tab 2816 \tab 9 \tab 11 \tab 30 \tab 1 \tab 2867 \cr
   Total \tab 7271 \tab 17 \tab 25 \tab 65 \tab 1 \tab 7379 \cr
}
}

\seealso{
Other Core: 
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
