% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{coden}
\alias{coden}
\title{Partner's denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
coden
}
\details{
Question What is your (PARTNER'S) religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion? A. What specific denomination is that, if any? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \verb{[1]} baptist
\item \verb{[2]} methodist
\item \verb{[3]} lutheran
\item \verb{[4]} presbyterian
\item \verb{[5]} episcopal
\item \verb{[6]} other denomination
\item \verb{[7]} no denomination or non-denominational church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6170/vshow}{GSS Data Explorer page for \code{coden}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab baptist \tab methodist \tab lutheran \tab presbyterian \tab episcopal \tab other denomination \tab no denomination or non-denominational church \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab - \tab 21 \tab 2 \tab 5 \tab 3 \tab 1 \tab 3 \tab 14 \tab 5 \tab 1920 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 25 \tab 6 \tab 2 \tab 1 \tab 1 \tab 10 \tab 15 \tab 4 \tab 2471 \tab 3 \tab - \tab 2538 \cr
   2016 \tab - \tab 30 \tab 8 \tab 8 \tab 4 \tab 6 \tab 14 \tab 23 \tab 5 \tab 2766 \tab 3 \tab - \tab 2867 \cr
   2018 \tab - \tab 26 \tab 4 \tab 7 \tab 1 \tab 1 \tab 15 \tab 18 \tab 4 \tab 2272 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 35 \tab 7 \tab 6 \tab 4 \tab 3 \tab 6 \tab 21 \tab 8 \tab 3942 \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 36 \tab 2 \tab 6 \tab 3 \tab 4 \tab 10 \tab 26 \tab - \tab 3457 \tab - \tab - \tab 3544 \cr
   Total \tab 0 \tab 173 \tab 29 \tab 34 \tab 16 \tab 16 \tab 58 \tab 117 \tab 26 \tab 16828 \tab 6 \tab 0 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
