% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{coden_1222}
\alias{coden_1222}
\title{Partner's denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
coden_1222
}
\details{
Question What is your (PARTNER'S) religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion? A. What specific denomination is that, if any?
As of 2024, CODEN has been recoded to only account for seven broad denomination categories. This has been retroactively applied to all previous years of CODEN. The original CODEN containing data from 2012 to 2022 has been renamed CODEN_1222.
}
\section{Values}{

\itemize{
\item \verb{[10]} american baptist association
\item \verb{[11]} american baptist churches in the u.s.a.
\item \verb{[12]} national baptist convention of america
\item \verb{[13]} national baptist convention u.s.a., inc.
\item \verb{[14]} southern baptist convention
\item \verb{[15]} other baptist churches
\item \verb{[18]} baptist, don't know which
\item \verb{[20]} african methodist episcopal church
\item \verb{[21]} african methodist episcopal zion church
\item \verb{[22]} united methodist church
\item \verb{[23]} other methodist churches
\item \verb{[28]} methodist, don't know which
\item \verb{[30]} american lutheran church
\item \verb{[31]} lutheran church in america
\item \verb{[32]} lutheran church - missouri synod
\item \verb{[33]} wisconsin evangelical lutheran synod
\item \verb{[34]} other lutheran churches
\item \verb{[35]} evangelical lutheran
\item \verb{[38]} lutheran, don't know which
\item \verb{[40]} presbyterian church in the u.s.
\item \verb{[41]} united presbyterian church in the u.s.a.
\item \verb{[42]} other presbyterian churches
\item \verb{[43]} presbyterian, merged
\item \verb{[48]} presbyterian, don't know which
\item \verb{[50]} episcopal church
\item \verb{[60]} other, specify
\item \verb{[70]} no denomination or non-denominational church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8226/vshow}{GSS Data Explorer page for \code{coden_1222}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab american baptist association \tab american baptist churches in the u.s.a. \tab national baptist convention of america \tab southern baptist convention \tab baptist, don't know which \tab united methodist church \tab methodist, don't know which \tab american lutheran church \tab lutheran church in america \tab evangelical lutheran \tab united presbyterian church in the u.s.a. \tab presbyterian, don't know which \tab episcopal church \tab other, specify \tab no denomination or non-denominational church \tab don't know \tab iap \tab lutheran, don't know which \tab no answer \tab other baptist churches \tab african methodist episcopal church \tab national baptist convention u.s.a., inc. \tab lutheran church - missouri synod \tab presbyterian church in the u.s. \tab presbyterian, merged \tab other lutheran churches \tab Total \cr
   2012 \tab - \tab 1 \tab 1 \tab 1 \tab 11 \tab 7 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 14 \tab 5 \tab 1920 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 2 \tab - \tab 8 \tab 15 \tab 6 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 10 \tab 15 \tab 4 \tab 2471 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 1 \tab 2 \tab 1 \tab 11 \tab 14 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 6 \tab 14 \tab 23 \tab 5 \tab 2766 \tab 7 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 3 \tab - \tab - \tab 11 \tab 10 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 15 \tab 18 \tab 4 \tab 2272 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 2 \tab 1 \tab 2 \tab 6 \tab 24 \tab 5 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 6 \tab 21 \tab 8 \tab 3942 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4032 \cr
   2022 \tab - \tab 5 \tab 1 \tab - \tab 6 \tab 24 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 10 \tab 26 \tab - \tab 3457 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3544 \cr
   Total \tab 0 \tab 12 \tab 7 \tab 4 \tab 53 \tab 94 \tab 20 \tab 7 \tab 6 \tab 3 \tab 3 \tab 3 \tab 10 \tab 16 \tab 58 \tab 117 \tab 26 \tab 16828 \tab 19 \tab 6 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{major1_1222}},
\code{\link{major2_1222}},
\code{\link{spden_7322}}

Other Religion: 
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{spden_7322}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
