% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{codipged}
\alias{codipged}
\title{Partner diploma, ged, or other}
\source{
General Social Survey https://gss.norc.org
}
\description{
codipged
}
\details{
Question Was that a diploma awarded by (RESPONDENT'S PARTNER)''s high school at graduation at the end of 12th grade, a GED awarded after (RESPONDENT''S PARTNER) took a test, or something else?
}
\section{Values}{

\itemize{
\item \code{1} high school diploma
\item \code{2} ged
\item \code{3} other
\item \code{5} high school diploma after post high school class
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6162/vshow}{GSS Data Explorer page for \code{codipged}}.
}

\seealso{
Other Core: 
\code{\link{cosector}},
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\keyword{variable}
