% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{coeduc}
\alias{coeduc}
\title{Highest year school completed, partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
coeduc
}
\details{
Question RESPONDENT'S PARTNER'S EDUCATION
}
\section{Values}{

\itemize{
\item \verb{[0]} no formal schooling
\item \verb{[1]} 1st grade
\item \verb{[2]} 2nd grade
\item \verb{[3]} 3rd grade
\item \verb{[4]} 4th grade
\item \verb{[5]} 5th grade
\item \verb{[6]} 6th grade
\item \verb{[7]} 7th grade
\item \verb{[8]} 8th grade
\item \verb{[9]} 9th grade
\item \verb{[10]} 10th grade
\item \verb{[11]} 11th grade
\item \verb{[12]} 12th grade
\item \verb{[13]} 1 year of college
\item \verb{[14]} 2 years of college
\item \verb{[15]} 3 years of college
\item \verb{[16]} 4 years of college
\item \verb{[17]} 5 years of college
\item \verb{[18]} 6 years of college
\item \verb{[19]} 7 years of college
\item \verb{[20]} 8 or more years of college
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6160/vshow}{GSS Data Explorer page for \code{coeduc}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 4th grade \tab 6th grade \tab 7th grade \tab 8th grade \tab 9th grade \tab 10th grade \tab 11th grade \tab 12th grade \tab 1 year of college \tab 2 years of college \tab 3 years of college \tab 4 years of college \tab 5 years of college \tab 6 years of college \tab 7 years of college \tab 8 or more years of college \tab don't know \tab no answer \tab 3rd grade \tab no formal schooling \tab 5th grade \tab Total \cr
   2012 \tab 1806 \tab 1 \tab 4 \tab 2 \tab 3 \tab 4 \tab 11 \tab 13 \tab 55 \tab 8 \tab 16 \tab 4 \tab 22 \tab 3 \tab 4 \tab 2 \tab 2 \tab 14 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2325 \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 6 \tab 13 \tab 88 \tab 15 \tab 18 \tab 10 \tab 21 \tab 4 \tab 6 \tab 1 \tab - \tab 14 \tab 4 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2603 \tab 1 \tab 4 \tab 1 \tab 4 \tab 8 \tab 9 \tab 16 \tab 100 \tab 24 \tab 22 \tab 8 \tab 30 \tab 10 \tab 4 \tab 4 \tab 8 \tab 7 \tab 3 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 2117 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 10 \tab 14 \tab 80 \tab 16 \tab 32 \tab 9 \tab 38 \tab 6 \tab 7 \tab 1 \tab 3 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 2348 \cr
   2021 \tab 3660 \tab 3 \tab 2 \tab 2 \tab 3 \tab 6 \tab 7 \tab 11 \tab 119 \tab 22 \tab 46 \tab 15 \tab 69 \tab 7 \tab 15 \tab 4 \tab 10 \tab 11 \tab 18 \tab 1 \tab 1 \tab - \tab 4032 \cr
   2022 \tab 3234 \tab - \tab 1 \tab 1 \tab 3 \tab 8 \tab 10 \tab 10 \tab 103 \tab 20 \tab 37 \tab 17 \tab 56 \tab 11 \tab 12 \tab 4 \tab 2 \tab 5 \tab 5 \tab 1 \tab 2 \tab 2 \tab 3544 \cr
   2024 \tab 3049 \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 12 \tab 7 \tab 89 \tab 21 \tab 27 \tab 8 \tab 51 \tab 7 \tab 12 \tab - \tab 6 \tab 9 \tab - \tab - \tab 1 \tab - \tab 3309 \cr
   Total \tab 18794 \tab 6 \tab 18 \tab 11 \tab 19 \tab 37 \tab 65 \tab 84 \tab 634 \tab 126 \tab 198 \tab 71 \tab 287 \tab 48 \tab 60 \tab 16 \tab 31 \tab 65 \tab 32 \tab 3 \tab 5 \tab 2 \tab 20612 \cr
}
}

\seealso{
Other Core: 
\code{\link{age}},
\code{\link{agekdbrn}},
\code{\link{codeg}},
\code{\link{degree}},
\code{\link{educ}},
\code{\link{maeduc}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{speduc}}
}
\concept{Core}
\keyword{variable}
