% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_04.R
\name{coevwork}
\alias{coevwork}
\title{Partner ever work as long as a year}
\source{
General Social Survey https://gss.norc.org
}
\description{
coevwork
}
\details{
Question Last week was your partner working full time, part time, going to school, keeping house, or what? C. IF RETIRED, IN SCHOOL, KEEPING HOUSE, OR OTHER: Did (he/she) ever work for as long as one year?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6154/vshow}{GSS Data Explorer page for \code{coevwork}}.
}

\seealso{
Other Core: 
\code{\link{cohrs1}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{divorce}},
\code{\link{spevwork}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spwrksta}},
\code{\link{widowed}}
}
\concept{Core}
\keyword{variable}
