% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{cofund}
\alias{cofund}
\title{How fundamentalist is partner's current religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
cofund
}
\details{
Question Fundamentalism/Liberalism of Partner's Religion
}
\section{Values}{

\itemize{
\item \code{1} fundamentalist
\item \code{2} moderate
\item \code{3} liberal
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6173/vshow}{GSS Data Explorer page for \code{cofund}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab fundamentalist \tab moderate \tab liberal \tab no answer \tab Total \cr
   2012 \tab 1807 \tab 23 \tab 65 \tab 64 \tab 15 \tab 1974 \cr
   2014 \tab 2326 \tab 30 \tab 82 \tab 79 \tab 21 \tab 2538 \cr
   2016 \tab 2603 \tab 33 \tab 102 \tab 100 \tab 29 \tab 2867 \cr
   2018 \tab 2117 \tab 38 \tab 77 \tab 95 \tab 21 \tab 2348 \cr
   2021 \tab 3681 \tab 41 \tab 95 \tab 192 \tab 23 \tab 4032 \cr
   2022 \tab 3230 \tab 48 \tab 95 \tab 147 \tab 24 \tab 3544 \cr
   Total \tab 15764 \tab 213 \tab 516 \tab 677 \tab 133 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
