% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{cogradtounder}
\alias{cogradtounder}
\title{Ratio of graduate to undergraduate students at college partner attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
cogradtounder
}
\details{
Question In what year did your partner receive the (COLLEGE NAME) degree?
L. Ratio of Graduate to Undergraduate students.
}
\section{Values}{

\itemize{
\item \verb{[1]} 0-.210 graduate student per undergraduate student
\item \verb{[2]} .211 or over
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6165/vshow}{GSS Data Explorer page for \code{cogradtounder}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 0-.210 graduate student per undergraduate student \tab .211 or over \tab no answer \tab Total \cr
   2012 \tab 1942 \tab 12 \tab 20 \tab - \tab 1974 \cr
   2014 \tab 2518 \tab 8 \tab 12 \tab - \tab 2538 \cr
   2016 \tab 2818 \tab 19 \tab 29 \tab 1 \tab 2867 \cr
   Total \tab 7278 \tab 39 \tab 61 \tab 1 \tab 7379 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
