% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{cohabfst}
\alias{cohabfst}
\title{People should live together before marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
cohabfst
}
\details{
Question Do you agree or disagree . . .
C. It's a good idea for a couple who intend to get married to live together first.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4407/vshow}{GSS Data Explorer page for \code{cohabfst}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 141 \tab 330 \tab 332 \tab 383 \tab 213 \tab 25 \tab 23 \tab 2992 \cr
   1998 \tab 1548 \tab 188 \tab 299 \tab 329 \tab 211 \tab 199 \tab 31 \tab 27 \tab 2832 \cr
   2002 \tab 1594 \tab 240 \tab 312 \tab 227 \tab 215 \tab 159 \tab 17 \tab 1 \tab 2765 \cr
   Total \tab 4687 \tab 569 \tab 941 \tab 888 \tab 809 \tab 571 \tab 73 \tab 51 \tab 8589 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabok}},
\code{\link{divbest}},
\code{\link{divifkd1}},
\code{\link{divnokd1}},
\code{\link{fewrksup}},
\code{\link{hubbywk1}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{singlpar}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
