% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{cohabit}
\alias{cohabit}
\title{Did r live with spouse before marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
cohabit
}
\details{
Question Did you live with your husband/wife before you got married?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4395/vshow}{GSS Data Explorer page for \code{cohabit}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1988 \tab 725 \tab 176 \tab 566 \tab 14 \tab 1481 \cr
   1994 \tab 2249 \tab 213 \tab 518 \tab 12 \tab 2992 \cr
   Total \tab 2974 \tab 389 \tab 1084 \tab 26 \tab 4473 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divhubby}},
\code{\link{earnsmor}},
\code{\link{evdiv}},
\code{\link{mawork14}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
