% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{cohrs2}
\alias{cohrs2}
\title{No. of hours partner usually works a week}
\source{
General Social Survey https://gss.norc.org
}
\description{
cohrs2
}
\details{
Question Last week was your (partner) working full time, part time, going to school, keeping house, or what?
B. IF WITH A JOB, BUT NOT AT WORK: How many hours a week does (he/she) usually work, at all jobs?
}
\section{Values}{

\itemize{
\item \verb{[89]} 89+ hours
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012, 2024 \cr
   A/B/C \tab 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6156/vshow}{GSS Data Explorer page for \code{cohrs2}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 30 \tab 42 \tab 40 \tab 50 \tab don't know \tab no answer \tab 34 \tab 52 \tab 32 \tab 25 \tab 35 \tab 70 \tab 89+ hours \tab Total \cr
   2012 \tab 1972 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2532 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2862 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2343 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4014 \tab 1 \tab - \tab 5 \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 4032 \cr
   2022 \tab 3538 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 17261 \tab 3 \tab 1 \tab 12 \tab 4 \tab 2 \tab 13 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spwrkslf}}
}
\concept{Core}
\keyword{variable}
