% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{colcom}
\alias{colcom}
\title{Should communist teacher be fired}
\source{
General Social Survey https://gss.norc.org
}
\description{
colcom
}
\details{
Question Now, I should like to ask you some questions about a man who admits he is a Communist.
B. Suppose he is teaching in a college. Should he be fired, or not?
}
\section{Values}{

\itemize{
\item \code{4} yes, fired
\item \code{5} not fired
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/256/vshow}{GSS Data Explorer page for \code{colcom}}.

Counts by year:\tabular{llllllll}{
   year \tab yes, fired \tab not fired \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 979 \tab 518 \tab 110 \tab 6 \tab - \tab - \tab 1613 \cr
   1973 \tab 823 \tab 585 \tab 88 \tab 8 \tab - \tab - \tab 1504 \cr
   1974 \tab 772 \tab 617 \tab 90 \tab 5 \tab - \tab - \tab 1484 \cr
   1976 \tab 803 \tab 620 \tab 73 \tab 3 \tab - \tab - \tab 1499 \cr
   1977 \tab 869 \tab 591 \tab 63 \tab 7 \tab - \tab - \tab 1530 \cr
   1980 \tab 782 \tab 595 \tab 88 \tab 3 \tab - \tab - \tab 1468 \cr
   1982 \tab 910 \tab 792 \tab 147 \tab 11 \tab - \tab - \tab 1860 \cr
   1984 \tab 717 \tab 671 \tab 77 \tab 8 \tab - \tab - \tab 1473 \cr
   1985 \tab 783 \tab 677 \tab 67 \tab 7 \tab - \tab - \tab 1534 \cr
   1987 \tab 860 \tab 837 \tab 115 \tab 7 \tab - \tab - \tab 1819 \cr
   1988 \tab 456 \tab 464 \tab 53 \tab 4 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 446 \tab 519 \tab 64 \tab 4 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 383 \tab 471 \tab 58 \tab 5 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 402 \tab 532 \tab 55 \tab 4 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 392 \tab 602 \tab 76 \tab 5 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 783 \tab 1093 \tab 113 \tab 7 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 714 \tab 1087 \tab 118 \tab 4 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 686 \tab 1059 \tab 130 \tab 7 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 679 \tab 1047 \tab 125 \tab 10 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 338 \tab 549 \tab 35 \tab 2 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 298 \tab 568 \tab 30 \tab 2 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 701 \tab 1202 \tab 94 \tab 6 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 500 \tab 783 \tab 65 \tab 4 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 437 \tab 776 \tab 58 \tab 10 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 409 \tab 804 \tab 81 \tab 14 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 540 \tab 1085 \tab 77 \tab 12 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 589 \tab 1198 \tab 89 \tab 14 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 510 \tab 955 \tab 97 \tab 12 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 383 \tab 910 \tab 7 \tab - \tab 2714 \tab 18 \tab 4032 \cr
   2022 \tab 379 \tab 719 \tab 50 \tab 8 \tab 2380 \tab 8 \tab 3544 \cr
   Total \tab 18323 \tab 22926 \tab 2393 \tab 199 \tab 22432 \tab 26 \tab 66299 \cr
}
}

\seealso{
Other Core: 
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Core}
\keyword{variable}
