% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{coldeg1}
\alias{coldeg1}
\title{The highest degree r have earned}
\source{
General Social Survey https://gss.norc.org
}
\description{
coldeg1
}
\details{
Question What is the highest degree you have earned?
}
\section{Values}{

\itemize{
\item \code{1} associate's
\item \code{2} bachelor's
\item \code{3} master's
\item \code{4} mba
\item \code{5} law
\item \code{6} phd
\item \code{7} md
\item \code{8} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3347/vshow}{GSS Data Explorer page for \code{coldeg1}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}

Other Education: 
\code{\link{majorcol}}
}
\concept{Education}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
