% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{colmajr2}
\alias{colmajr2}
\title{R 2nd major in college}
\source{
General Social Survey https://gss.norc.org
}
\description{
colmajr2
}
\details{
Question A. Could you tell me your major field of study (while you were) in college?
}
\section{Values}{

\itemize{
\item \verb{[0]} none
\item \verb{[1]} english, literature
\item \verb{[2]} foreign language
\item \verb{[3]} fine arts
\item \verb{[4]} other humanities
\item \verb{[5]} economics
\item \verb{[6]} psychology
\item \verb{[7]} government-civics
\item \verb{[8]} social science-other social study
\item \verb{[9]} mathematics
\item \verb{[10]} science
\item \verb{[11]} other
\item \verb{[12]} business
\item \verb{[13]} education
\item \verb{[14]} engineering
\item \verb{[15]} other vocational
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1451/vshow}{GSS Data Explorer page for \code{colmajr2}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab foreign language \tab fine arts \tab other humanities \tab economics \tab psychology \tab social science-other social study \tab mathematics \tab science \tab other \tab business \tab education \tab engineering \tab other vocational \tab no answer \tab not available in this year \tab Total \cr
   1993 \tab 1507 \tab 2 \tab 2 \tab 1 \tab 1 \tab 5 \tab 5 \tab 6 \tab 5 \tab 16 \tab 15 \tab 17 \tab 6 \tab 10 \tab 8 \tab - \tab 1606 \cr
   Total \tab 1507 \tab 2 \tab 2 \tab 1 \tab 1 \tab 5 \tab 5 \tab 6 \tab 5 \tab 16 \tab 15 \tab 17 \tab 6 \tab 10 \tab 8 \tab 0 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{bigband}},
\code{\link{blues}},
\code{\link{blugrass}},
\code{\link{camping}},
\code{\link{classicl}},
\code{\link{classics}},
\code{\link{colmajr1}},
\code{\link{conrock}},
\code{\link{country}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{english}},
\code{\link{excelart}},
\code{\link{folk}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{gospel}},
\code{\link{grtbooks}},
\code{\link{hosthome}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{huntfish}},
\code{\link{hvymetal}},
\code{\link{impcultr}},
\code{\link{impfinan}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{jazz}},
\code{\link{judgeart}},
\code{\link{latin}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{makeart}},
\code{\link{modpaint}},
\code{\link{moodeasy}},
\code{\link{mostimp1}},
\code{\link{mostimp2}},
\code{\link{mostimp3}},
\code{\link{mostimp4}},
\code{\link{musicals}},
\code{\link{newage}},
\code{\link{obtohelp}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{opoutcme}},
\code{\link{pclit}},
\code{\link{perform}},
\code{\link{plymusic}},
\code{\link{popmusic}},
\code{\link{rap}},
\code{\link{readfict}},
\code{\link{reggae}},
\code{\link{richpoor}},
\code{\link{seemovie}},
\code{\link{selfirst}},
\code{\link{standup}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}},
\code{\link{united}},
\code{\link{usevcr}},
\code{\link{visitart}}

Other Education: 
\code{\link{actgreek}},
\code{\link{actschl}},
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{bilinged}},
\code{\link{cathsch}},
\code{\link{chldeduc}},
\code{\link{chrissch}},
\code{\link{classics}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{coldeg1}},
\code{\link{colmajr1}},
\code{\link{colop}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{datesch}},
\code{\link{degree}},
\code{\link{denomsch}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{educ}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{educemp}},
\code{\link{educop}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{govedop}},
\code{\link{homesch}},
\code{\link{hsbasics}},
\code{\link{hsbio}},
\code{\link{hscaring}},
\code{\link{hschem}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsmath}},
\code{\link{hsorder}},
\code{\link{hsphys}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inteduc}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{majorcol}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memschl}},
\code{\link{natdeg}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{opeduc}},
\code{\link{oppared}},
\code{\link{othsch}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{relsch}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{schleth}},
\code{\link{schlhead}},
\code{\link{schoolng}},
\code{\link{scimath}},
\code{\link{sexeduc}},
\code{\link{solgreek}},
\code{\link{solschl}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{speduc}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{teacheth}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{trstprof}},
\code{\link{undoccol}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{whoteach}}
}
\concept{Culture}
\concept{Education}
\keyword{variable}
