% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{colmil}
\alias{colmil}
\title{Allow militarist to teach}
\source{
General Social Survey https://gss.norc.org
}
\description{
colmil
}
\details{
Question Consider a person who advocates doing away with elections and letting the military run the country.
B. Should such a person be allowed to teach in a college or university, or not?
}
\section{Values}{

\itemize{
\item \code{4} yes, allowed to teach
\item \code{5} not allowed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/259/vshow}{GSS Data Explorer page for \code{colmil}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, allowed to teach \tab not allowed \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1976 \tab - \tab 553 \tab 886 \tab 55 \tab 5 \tab - \tab 1499 \cr
   1977 \tab - \tab 519 \tab 968 \tab 39 \tab 4 \tab - \tab 1530 \cr
   1980 \tab - \tab 577 \tab 827 \tab 60 \tab 4 \tab - \tab 1468 \cr
   1982 \tab - \tab 679 \tab 1063 \tab 106 \tab 12 \tab - \tab 1860 \cr
   1984 \tab - \tab 604 \tab 819 \tab 44 \tab 6 \tab - \tab 1473 \cr
   1985 \tab - \tab 608 \tab 871 \tab 49 \tab 6 \tab - \tab 1534 \cr
   1987 \tab - \tab 693 \tab 1038 \tab 82 \tab 6 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 362 \tab 575 \tab 34 \tab 6 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 415 \tab 562 \tab 52 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 394 \tab 481 \tab 38 \tab 4 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 424 \tab 513 \tab 51 \tab 5 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 510 \tab 514 \tab 47 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 910 \tab 1010 \tab 71 \tab 5 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 925 \tab 900 \tab 94 \tab 4 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 966 \tab 820 \tab 92 \tab 4 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 890 \tab 879 \tab 84 \tab 8 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 510 \tab 391 \tab 22 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 477 \tab 398 \tab 20 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 1051 \tab 892 \tab 55 \tab 5 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 682 \tab 617 \tab 50 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 726 \tab 519 \tab 31 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 740 \tab 529 \tab 36 \tab 3 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 994 \tab 670 \tab 39 \tab 11 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 1104 \tab 725 \tab 53 \tab 8 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 888 \tab 637 \tab 43 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 1357 \tab 1243 \tab 1394 \tab 6 \tab - \tab 32 \tab 4032 \cr
   Total \tab 18695 \tab 18444 \tab 19498 \tab 1353 \tab 132 \tab 32 \tab 58154 \cr
}
}

\seealso{
Other Core: 
\code{\link{colcom}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Core}
\keyword{variable}
