% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_257.R
\name{comamind}
\alias{comamind}
\title{Percentage american indian living in r's community}
\source{
General Social Survey https://gss.norc.org
}
\description{
comamind
}
\details{
Question Just your best guess–what percentage of the people who live in your local community is each group?
F. American Indians
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2328/vshow}{GSS Data Explorer page for \code{comamind}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 15 \tab 2 \tab 20 \tab 21 \tab 25 \tab 26 \tab 3 \tab 30 \tab 35 \tab 4 \tab 40 \tab 45 \tab 49 \tab 5 \tab 50 \tab 55 \tab 6 \tab 60 \tab 75 \tab 8 \tab 80 \tab 90 \tab 95 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 521 \tab 212 \tab 98 \tab 1 \tab 2 \tab 14 \tab 86 \tab 35 \tab 1 \tab 10 \tab 1 \tab 21 \tab 18 \tab 2 \tab 7 \tab 5 \tab 2 \tab 1 \tab 173 \tab 9 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 97 \tab 66 \tab 2817 \cr
   Total \tab 1419 \tab 521 \tab 212 \tab 98 \tab 1 \tab 2 \tab 14 \tab 86 \tab 35 \tab 1 \tab 10 \tab 1 \tab 21 \tab 18 \tab 2 \tab 7 \tab 5 \tab 2 \tab 1 \tab 173 \tab 9 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 97 \tab 66 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{ethchng}},
\code{\link{hispwork}},
\code{\link{knwwht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
