% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_257.R
\name{comjews}
\alias{comjews}
\title{Percentage jews living in r's community}
\source{
General Social Survey https://gss.norc.org
}
\description{
comjews
}
\details{
Question Just your best guess–what percentage of the people who live in your local community is each group?
C. Jews?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2325/vshow}{GSS Data Explorer page for \code{comjews}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 15 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 33 \tab 35 \tab 4 \tab 40 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 70 \tab 75 \tab 8 \tab 80 \tab 90 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 375 \tab 125 \tab 138 \tab 2 \tab 26 \tab 94 \tab 50 \tab 16 \tab 34 \tab 31 \tab 1 \tab 8 \tab 8 \tab 12 \tab 214 \tab 16 \tab 3 \tab 5 \tab 3 \tab 2 \tab 1 \tab 7 \tab 2 \tab 1 \tab 149 \tab 75 \tab 2817 \cr
   Total \tab 1419 \tab 375 \tab 125 \tab 138 \tab 2 \tab 26 \tab 94 \tab 50 \tab 16 \tab 34 \tab 31 \tab 1 \tab 8 \tab 8 \tab 12 \tab 214 \tab 16 \tab 3 \tab 5 \tab 3 \tab 2 \tab 1 \tab 7 \tab 2 \tab 1 \tab 149 \tab 75 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comwht}},
\code{\link{ethchng}},
\code{\link{hispwork}},
\code{\link{knwwht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
