% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{commun10}
\alias{commun10}
\title{Feelings about communism 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
commun10
}
\details{
Question What about ten years ago, that is in 1972, which of these statements came closest to how you felt about Communism as a form of government?  CODE ONE.
}
\section{Values}{

\itemize{
\item \verb{[1]} worst kind
\item \verb{[2]} bad, not worst
\item \verb{[3]} ok for some
\item \verb{[4]} good form
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5489/vshow}{GSS Data Explorer page for \code{commun10}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab worst kind \tab bad, not worst \tab ok for some \tab good form \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 1090 \tab 327 \tab 220 \tab 35 \tab 179 \tab 9 \tab 1860 \cr
   Total \tab 0 \tab 1090 \tab 327 \tab 220 \tab 35 \tab 179 \tab 9 \tab 1860 \cr
}
}

\seealso{
Other War and International: 
\code{\link{commun}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}
}
\concept{War and International}
\keyword{variable}
