% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_02.R
\name{commute}
\alias{commute}
\title{Travel time to work}
\source{
General Social Survey https://gss.norc.org
}
\description{
commute
}
\details{
Question A. What kind of work do you (did you normally) do?  That is, what (is/was) your job called?
OCCUPATION:  (See REMARKS)
B. What (do/did) you actually do in that job?  Tell me, what (are/were) some of your main duties?
C. What kind of place (do/did) you work for?
INDUSTRY:  (See REMARKS)
D. What (do/did) they (make/do)?
E. (Are/Were) you self employed or (do/did) you work for someone else?
F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)?
G. About how much time (does/did) it usually take you to travel to work - about how many minutes?
}
\section{Values}{

\itemize{
\item \verb{[97]} 97+ minutes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/11/vshow}{GSS Data Explorer page for \code{commute}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 2 \tab 20 \tab 22 \tab 24 \tab 25 \tab 28 \tab 3 \tab 30 \tab 32 \tab 35 \tab 4 \tab 40 \tab 43 \tab 45 \tab 46 \tab 5 \tab 50 \tab 55 \tab 6 \tab 60 \tab 7 \tab 70 \tab 75 \tab 8 \tab 9 \tab 90 \tab 97+ minutes \tab don't know \tab no answer \tab Total \cr
   1986 \tab 100 \tab 54 \tab 14 \tab 190 \tab 1 \tab 16 \tab 1 \tab 2 \tab 203 \tab 1 \tab 3 \tab 34 \tab 180 \tab 2 \tab 1 \tab 60 \tab 1 \tab 23 \tab 155 \tab 2 \tab 30 \tab 6 \tab 22 \tab 1 \tab 63 \tab 1 \tab 153 \tab 5 \tab 3 \tab 2 \tab 55 \tab 13 \tab 4 \tab 4 \tab 8 \tab 1 \tab 8 \tab 8 \tab 8 \tab 32 \tab 1470 \cr
   Total \tab 100 \tab 54 \tab 14 \tab 190 \tab 1 \tab 16 \tab 1 \tab 2 \tab 203 \tab 1 \tab 3 \tab 34 \tab 180 \tab 2 \tab 1 \tab 60 \tab 1 \tab 23 \tab 155 \tab 2 \tab 30 \tab 6 \tab 22 \tab 1 \tab 63 \tab 1 \tab 153 \tab 5 \tab 3 \tab 2 \tab 55 \tab 13 \tab 4 \tab 4 \tab 8 \tab 1 \tab 8 \tab 8 \tab 8 \tab 32 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{found}},
\code{\link{indus07}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{prestg80}}
}
\concept{Core}
\keyword{variable}
