% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_379.R
\name{comorsci}
\alias{comorsci}
\title{Antarctica should be reserved or open}
\source{
General Social Survey https://gss.norc.org
}
\description{
comorsci
}
\details{
Question Some people think that Antarctica should be reserved primarily for scientific purposes. Others think it should be open for tourism, fishing, exploration for oil, and other commercial purposes. Please look at Card B23. Which statement best describes your view on this issue?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly support reserving antarctica
\item \verb{[2]} somewhat support reserving antarctica
\item \verb{[3]} dont lean one way or the other on this issue
\item \verb{[4]} somewhat support opening antarctica
\item \verb{[5]} strongly support opening antarctica
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3437/vshow}{GSS Data Explorer page for \code{comorsci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly support reserving antarctica \tab somewhat support reserving antarctica \tab dont lean one way or the other on this issue \tab somewhat support opening antarctica \tab strongly support opening antarctica \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 436 \tab 435 \tab 597 \tab 252 \tab 83 \tab 59 \tab 2 \tab 4510 \cr
   2010 \tab 1338 \tab 148 \tab 173 \tab 229 \tab 89 \tab 41 \tab 14 \tab 12 \tab 2044 \cr
   Total \tab 3984 \tab 584 \tab 608 \tab 826 \tab 341 \tab 124 \tab 73 \tab 14 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{extinct}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{nosun}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
