% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_572.R
\name{compfin}
\alias{compfin}
\title{Govt check personal financial records}
\source{
General Social Survey https://gss.norc.org
}
\description{
compfin
}
\details{
Question Protecting secrets is a continuing concern of the government. I am going to ask you about various security measures that the government might apply to individuals with a SECRET or TOP SECRET clearance. Do you (strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree) that people with SECRET or TOP SECRET clearance should be subject to:
A. Computer checks of personal financial records
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5249/vshow}{GSS Data Explorer page for \code{compfin}}.
}

\seealso{
Other National Security: 
\code{\link{bugging}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{lietest}},
\code{\link{testdrug}}

Other United States: 
\code{\link{bugging}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{lietest}},
\code{\link{testdrug}}
}
\concept{National Security}
\concept{United States}
\keyword{variable}
