% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{compfri}
\alias{compfri}
\title{Does r use computer at friends house}
\source{
General Social Survey https://gss.norc.org
}
\description{
compfri
}
\details{
Question D. Friend's house?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2582/vshow}{GSS Data Explorer page for \code{compfri}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2665 \tab 71 \tab 48 \tab 33 \tab 2817 \cr
   2002 \tab 2537 \tab 150 \tab 75 \tab 3 \tab 2765 \cr
   Total \tab 5202 \tab 221 \tab 123 \tab 36 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Information Society}
\keyword{variable}
