% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{comphome}
\alias{comphome}
\title{Does r have computer at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
comphome
}
\details{
Question A. Do you have one or more computers at your home? (INTERVIEWER: INCLUDE LAPTOPS THAT R CAN USE AT WORK OR HOME)
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2549/vshow}{GSS Data Explorer page for \code{comphome}}.
}

\seealso{
Other Information Society: 
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
