% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{complib}
\alias{complib}
\title{Does r use computer at library}
\source{
General Social Survey https://gss.norc.org
}
\description{
complib
}
\details{
Question C. Library?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2581/vshow}{GSS Data Explorer page for \code{complib}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2665 \tab 65 \tab 53 \tab 34 \tab 2817 \cr
   2002 \tab 2537 \tab 107 \tab 118 \tab 3 \tab 2765 \cr
   Total \tab 5202 \tab 172 \tab 171 \tab 37 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Information Society}
\keyword{variable}
