% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_571.R
\name{comporn}
\alias{comporn}
\title{Punishment for downloading porn at office}
\source{
General Social Survey https://gss.norc.org
}
\description{
comporn
}
\details{
Question Now I’m going to ask you about several offenses involving the use of a government computer system. For each case, indicate what you would consider to be the appropriate punishment when the offender is a military service member or government employee.
F. Downloading pornographic material from the Internet to a computer at the office.
}
\section{Values}{

\itemize{
\item \code{1} reprimand
\item \code{2} dismissal from the military/fired by the government
\item \code{3} one year in prison
\item \code{4} 5-9 years in prison
\item \code{5} 10-20 years in prison
\item \code{6} life in prison
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5242/vshow}{GSS Data Explorer page for \code{comporn}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab reprimand \tab dismissal from the military/fired by the government \tab one year in prison \tab 5-9 years in prison \tab 10-20 years in prison \tab life in prison \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 386 \tab 545 \tab 138 \tab 134 \tab 76 \tab 61 \tab 64 \tab 15 \tab 2817 \cr
   Total \tab 1398 \tab 386 \tab 545 \tab 138 \tab 134 \tab 76 \tab 61 \tab 64 \tab 15 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}}
}
\concept{National Security}
\keyword{variable}
