% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{comprend}
\alias{comprend}
\title{R's understanding of questions}
\source{
General Social Survey https://gss.norc.org
}
\description{
comprend
}
\details{
Question Was respondent's understanding of the questions . . . (CODE ONE.)
}
\section{Values}{

\itemize{
\item \verb{[1]} good
\item \verb{[2]} fair
\item \verb{[3]} poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/696/vshow}{GSS Data Explorer page for \code{comprend}}.

Counts by year:\tabular{llllllll}{
   year \tab good \tab fair \tab poor \tab no answer \tab don't know \tab iap \tab Total \cr
   1972 \tab 1237 \tab 312 \tab 50 \tab 14 \tab - \tab - \tab 1613 \cr
   1973 \tab 1170 \tab 280 \tab 44 \tab 10 \tab - \tab - \tab 1504 \cr
   1974 \tab 1163 \tab 278 \tab 38 \tab 5 \tab - \tab - \tab 1484 \cr
   1975 \tab 1187 \tab 251 \tab 44 \tab 8 \tab - \tab - \tab 1490 \cr
   1976 \tab 1158 \tab 280 \tab 43 \tab 18 \tab - \tab - \tab 1499 \cr
   1977 \tab 1220 \tab 247 \tab 55 \tab 8 \tab - \tab - \tab 1530 \cr
   1978 \tab 1158 \tab 299 \tab 53 \tab 22 \tab - \tab - \tab 1532 \cr
   1980 \tab 1120 \tab 269 \tab 51 \tab 28 \tab - \tab - \tab 1468 \cr
   1982 \tab 1409 \tab 360 \tab 80 \tab 11 \tab - \tab - \tab 1860 \cr
   1983 \tab 1270 \tab 255 \tab 39 \tab 35 \tab - \tab - \tab 1599 \cr
   1984 \tab 1150 \tab 230 \tab 63 \tab 30 \tab - \tab - \tab 1473 \cr
   1985 \tab 1267 \tab 204 \tab 39 \tab 24 \tab - \tab - \tab 1534 \cr
   1986 \tab 1145 \tab 234 \tab 47 \tab 44 \tab - \tab - \tab 1470 \cr
   1987 \tab 1390 \tab 311 \tab 95 \tab 23 \tab - \tab - \tab 1819 \cr
   1988 \tab 1145 \tab 258 \tab 58 \tab 20 \tab - \tab - \tab 1481 \cr
   1989 \tab 1197 \tab 249 \tab 58 \tab 33 \tab - \tab - \tab 1537 \cr
   1990 \tab 1083 \tab 210 \tab 60 \tab 19 \tab - \tab - \tab 1372 \cr
   1991 \tab 1233 \tab 207 \tab 48 \tab 29 \tab - \tab - \tab 1517 \cr
   1993 \tab 1310 \tab 240 \tab 49 \tab 7 \tab - \tab - \tab 1606 \cr
   1994 \tab 2480 \tab 423 \tab 71 \tab 18 \tab - \tab - \tab 2992 \cr
   1996 \tab 2411 \tab 391 \tab 73 \tab 29 \tab - \tab - \tab 2904 \cr
   1998 \tab 2377 \tab 360 \tab 67 \tab 26 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 2309 \tab 391 \tab 68 \tab 49 \tab - \tab - \tab 2817 \cr
   2002 \tab 2380 \tab 303 \tab 69 \tab - \tab 13 \tab - \tab 2765 \cr
   2004 \tab 2450 \tab 312 \tab 39 \tab 11 \tab - \tab - \tab 2812 \cr
   2006 \tab 3893 \tab 516 \tab 94 \tab 7 \tab - \tab - \tab 4510 \cr
   2008 \tab 1674 \tab 286 \tab 60 \tab 3 \tab - \tab - \tab 2023 \cr
   2010 \tab 1633 \tab 341 \tab 57 \tab 9 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 1610 \tab 298 \tab 55 \tab 4 \tab 7 \tab - \tab 1974 \cr
   2014 \tab 2016 \tab 462 \tab 52 \tab 6 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2390 \tab 417 \tab 50 \tab 8 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 1957 \tab 331 \tab 48 \tab 9 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 254 \tab 66 \tab 5 \tab - \tab - \tab 3707 \tab 4032 \cr
   2022 \tab 1461 \tab 373 \tab 53 \tab - \tab 15 \tab 1642 \tab 3544 \cr
   Total \tab 54307 \tab 10244 \tab 1875 \tab 567 \tab 48 \tab 5349 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abspno}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{form}},
\code{\link{phone}},
\code{\link{radiohrs}}
}
\concept{Core}
\keyword{variable}
