% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{compschl}
\alias{compschl}
\title{Does r use computer at school}
\source{
General Social Survey https://gss.norc.org
}
\description{
compschl
}
\details{
Question B. School?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2580/vshow}{GSS Data Explorer page for \code{compschl}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2665 \tab 39 \tab 85 \tab 28 \tab 2817 \cr
   2002 \tab 2537 \tab 69 \tab 156 \tab 3 \tab 2765 \cr
   Total \tab 5202 \tab 108 \tab 241 \tab 31 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Information Society}
\keyword{variable}
