% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{compwork}
\alias{compwork}
\title{Does r use computer at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
compwork
}
\details{
Question Do you use a computer at work–that is, at a workplace away from your home?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2566/vshow}{GSS Data Explorer page for \code{compwork}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2294 \tab 291 \tab 192 \tab 40 \tab - \tab 2817 \cr
   2002 \tab 855 \tab 1177 \tab 716 \tab 16 \tab 1 \tab 2765 \cr
   Total \tab 3149 \tab 1468 \tab 908 \tab 56 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emhrh}},
\code{\link{emminh}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{perhrh}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwhrh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
