% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{comtype}
\alias{comtype}
\title{Type of community in which r lives}
\source{
General Social Survey https://gss.norc.org
}
\description{
comtype
}
\details{
Question Would you describe the place where you live as…
}
\section{Values}{

\itemize{
\item \verb{[1]} big city
\item \verb{[2]} suburbs, outskirts
\item \verb{[3]} small town
\item \verb{[4]} country village
\item \verb{[5]} farm, cntry home
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4777/vshow}{GSS Data Explorer page for \code{comtype}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab big city \tab suburbs, outskirts \tab small town \tab country village \tab farm, cntry home \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1993 \tab 49 \tab 266 \tab 406 \tab 590 \tab 66 \tab 192 \tab 1 \tab 36 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 268 \tab 328 \tab 552 \tab 61 \tab 145 \tab 1 \tab 31 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 238 \tab 292 \tab 499 \tab 51 \tab 127 \tab - \tab 69 \tab - \tab 2817 \cr
   Total \tab 3196 \tab 772 \tab 1026 \tab 1641 \tab 178 \tab 464 \tab 2 \tab 136 \tab 0 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{anrights}},
\code{\link{antests}},
\code{\link{bizpeop}},
\code{\link{busdecid}},
\code{\link{busgrn}},
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{chemfam}},
\code{\link{chemfree}},
\code{\link{chemgen}},
\code{\link{drivless}},
\code{\link{econgrn}},
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{excldimm}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grncon}},
\code{\link{grndemo}},
\code{\link{grnecon}},
\code{\link{grneffme}},
\code{\link{grnexagg}},
\code{\link{grngroup}},
\code{\link{grnintl}},
\code{\link{grnmoney}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grnsign}},
\code{\link{grnsol}},
\code{\link{grntaxes}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{h2oless}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{helpharm}},
\code{\link{ihlpgrn}},
\code{\link{impgrn}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{ldcgrn}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{nobuygrn}},
\code{\link{nomeat}},
\code{\link{nukeacc}},
\code{\link{nukefam}},
\code{\link{nukegen}},
\code{\link{obeythnk}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{peopgrn}},
\code{\link{popgrwth}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{pubdecid}},
\code{\link{recycle}},
\code{\link{redcehme}},
\code{\link{resnatur}},
\code{\link{scifaith}},
\code{\link{scigrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{sciworse}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{tempgen1}},
\code{\link{toodifme}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}},
\code{\link{usdoenuf}},
\code{\link{waterfam}},
\code{\link{watergen}}

Other Residence: 
\code{\link{abroad}},
\code{\link{agecmeus}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{comamind}},
\code{\link{comyear}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwell5}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{geomobil}},
\code{\link{hspasn10}},
\code{\link{liveasns}},
\code{\link{livecom}},
\code{\link{livecom1}},
\code{\link{livehome}},
\code{\link{livehsps}},
\code{\link{loclived}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{res16}},
\code{\link{res161}},
\code{\link{res2006}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{ISSP Environment}
\concept{Residence}
\keyword{variable}
