% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{conage}
\alias{conage}
\title{Conflict between young and old}
\source{
General Social Survey https://gss.norc.org
}
\description{
conage
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
G. Young people and older people?
}
\section{Values}{

\itemize{
\item \code{1} very strong conflicts
\item \code{2} strong conflicts
\item \code{3} not very strong conflicts
\item \code{4} there are no conflicts
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4312/vshow}{GSS Data Explorer page for \code{conage}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\concept{Social Class}
\keyword{variable}
