% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{conbiz}
\alias{conbiz}
\title{Confidence in business and industry}
\source{
General Social Survey https://gss.norc.org
}
\description{
conbiz
}
\details{
Question How much confidence do you have in. . .
B. Business and industry
}
\section{Values}{

\itemize{
\item \verb{[1]} complete confidence
\item \verb{[2]} a great deal of confidence
\item \verb{[3]} some confidence
\item \verb{[4]} very little confidence
\item \verb{[5]} no confidence at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4640/vshow}{GSS Data Explorer page for \code{conbiz}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab complete confidence \tab a great deal of confidence \tab some confidence \tab very little confidence \tab no confidence at all \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab 158 \tab 44 \tab 241 \tab 714 \tab 220 \tab 54 \tab 40 \tab 46 \tab - \tab 1517 \cr
   1998 \tab 1548 \tab 34 \tab 271 \tab 659 \tab 186 \tab 45 \tab 45 \tab 44 \tab - \tab 2832 \cr
   2008 \tab 658 \tab 35 \tab 199 \tab 803 \tab 231 \tab 78 \tab 19 \tab - \tab - \tab 2023 \cr
   2018 \tab 1173 \tab 35 \tab 223 \tab 687 \tab 166 \tab 41 \tab 15 \tab 8 \tab - \tab 2348 \cr
   Total \tab 3537 \tab 148 \tab 934 \tab 2863 \tab 803 \tab 218 \tab 119 \tab 98 \tab 0 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{befair}},
\code{\link{bible1}},
\code{\link{cantrust}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{hapunhap}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{madenkid}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relcand}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}

Other Confidence: 
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{congovt}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{conschls}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{govdook}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{trustsci}}
}
\concept{Confidence}
\concept{ISSP Religion}
\keyword{variable}
