% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{conchurh}
\alias{conchurh}
\title{Confidence in churches and religious orgs}
\source{
General Social Survey https://gss.norc.org
}
\description{
conchurh
}
\details{
Question How much confidence do you have in. . .
D. Churches and religious organizations
}
\section{Values}{

\itemize{
\item \verb{[1]} complete confidence
\item \verb{[2]} a great deal of confidence
\item \verb{[3]} some confidence
\item \verb{[4]} very little confidence
\item \verb{[5]} no confidence at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4642/vshow}{GSS Data Explorer page for \code{conchurh}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab complete confidence \tab a great deal of confidence \tab some confidence \tab very little confidence \tab no confidence at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 165 \tab 357 \tab 524 \tab 161 \tab 82 \tab 32 \tab 38 \tab 1517 \cr
   1998 \tab 1548 \tab 115 \tab 409 \tab 482 \tab 134 \tab 64 \tab 41 \tab 39 \tab 2832 \cr
   2008 \tab 658 \tab 90 \tab 293 \tab 679 \tab 177 \tab 107 \tab 13 \tab 6 \tab 2023 \cr
   2018 \tab 1173 \tab 73 \tab 288 \tab 532 \tab 179 \tab 77 \tab 17 \tab 9 \tab 2348 \cr
   Total \tab 3537 \tab 443 \tab 1347 \tab 2217 \tab 651 \tab 330 \tab 103 \tab 92 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
