% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{conclerg}
\alias{conclerg}
\title{Confidence in organized religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
conclerg
}
\details{
Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?
READ EACH ITEM; CODE ONE FOR EACH.
C. Organized religion
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} only some
\item \verb{[3]} hardly any
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/450/vshow}{GSS Data Explorer page for \code{conclerg}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab only some \tab hardly any \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1973 \tab - \tab 520 \tab 685 \tab 237 \tab 53 \tab 9 \tab - \tab 1504 \cr
   1974 \tab - \tab 656 \tab 634 \tab 160 \tab 31 \tab 3 \tab - \tab 1484 \cr
   1975 \tab - \tab 362 \tab 711 \tab 317 \tab 95 \tab 5 \tab - \tab 1490 \cr
   1976 \tab - \tab 457 \tab 667 \tab 273 \tab 94 \tab 8 \tab - \tab 1499 \cr
   1977 \tab - \tab 610 \tab 688 \tab 177 \tab 51 \tab 4 \tab - \tab 1530 \cr
   1978 \tab - \tab 468 \tab 722 \tab 278 \tab 58 \tab 6 \tab - \tab 1532 \cr
   1980 \tab - \tab 517 \tab 631 \tab 261 \tab 56 \tab 3 \tab - \tab 1468 \cr
   1982 \tab - \tab 602 \tab 910 \tab 281 \tab 60 \tab 7 \tab - \tab 1860 \cr
   1983 \tab - \tab 450 \tab 805 \tab 277 \tab 61 \tab 6 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 304 \tab 457 \tab 184 \tab 34 \tab 10 \tab - \tab 1473 \cr
   1986 \tab - \tab 369 \tab 740 \tab 308 \tab 50 \tab 3 \tab - \tab 1470 \cr
   1987 \tab - \tab 534 \tab 881 \tab 339 \tab 57 \tab 8 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 199 \tab 460 \tab 308 \tab 28 \tab 2 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 223 \tab 459 \tab 313 \tab 36 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 203 \tab 442 \tab 212 \tab 41 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 256 \tab 523 \tab 211 \tab 25 \tab 2 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 240 \tab 521 \tab 257 \tab 33 \tab 6 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 486 \tab 1035 \tab 436 \tab 45 \tab 9 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 482 \tab 984 \tab 365 \tab 91 \tab 3 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 512 \tab 974 \tab 358 \tab 61 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 519 \tab 934 \tab 356 \tab 81 \tab 6 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 168 \tab 511 \tab 217 \tab 14 \tab 2 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 206 \tab 442 \tab 208 \tab 17 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 487 \tab 1014 \tab 429 \tab 54 \tab 5 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 272 \tab 722 \tab 342 \tab 24 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 271 \tab 720 \tab 328 \tab 49 \tab 9 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 276 \tab 686 \tab 323 \tab 41 \tab 12 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 338 \tab 896 \tab 402 \tab 34 \tab 17 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 379 \tab 1032 \tab 503 \tab 30 \tab 12 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 315 \tab 806 \tab 392 \tab 38 \tab 12 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 394 \tab 1377 \tab 879 \tab 5 \tab 1 \tab 16 \tab 4032 \cr
   2022 \tab 1173 \tab 336 \tab 1157 \tab 813 \tab 43 \tab 10 \tab 12 \tab 3544 \cr
   Total \tab 20145 \tab 12411 \tab 24226 \tab 10744 \tab 1490 \tab 199 \tab 28 \tab 69243 \cr
}
}

\seealso{
Other Core: 
\code{\link{conbus}},
\code{\link{coneduc}},
\code{\link{confed}},
\code{\link{confinan}},
\code{\link{conlabor}},
\code{\link{conpress}},
\code{\link{satfam}},
\code{\link{satfrnd}},
\code{\link{sathealt}}
}
\concept{Core}
\keyword{variable}
