% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{concong}
\alias{concong}
\title{Confidence in us congress}
\source{
General Social Survey https://gss.norc.org
}
\description{
concong
}
\details{
Question How much confidence do you have in. . .
A. U.S. Congress
}
\section{Values}{

\itemize{
\item \verb{[1]} complete confidence
\item \verb{[2]} a great deal of confidence
\item \verb{[3]} some confidence
\item \verb{[4]} very little confidence
\item \verb{[5]} no confidence at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4639/vshow}{GSS Data Explorer page for \code{concong}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab complete confidence \tab a great deal of confidence \tab some confidence \tab very little confidence \tab no confidence at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 102 \tab 236 \tab 594 \tab 264 \tab 91 \tab 44 \tab 28 \tab 1517 \cr
   1998 \tab 1548 \tab 28 \tab 130 \tab 638 \tab 319 \tab 93 \tab 41 \tab 35 \tab 2832 \cr
   2008 \tab 658 \tab 32 \tab 105 \tab 652 \tab 364 \tab 187 \tab 23 \tab 2 \tab 2023 \cr
   2018 \tab 1173 \tab 24 \tab 57 \tab 438 \tab 458 \tab 166 \tab 23 \tab 9 \tab 2348 \cr
   Total \tab 3537 \tab 186 \tab 528 \tab 2322 \tab 1405 \tab 537 \tab 131 \tab 74 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}
}
\concept{ISSP Religion}
\keyword{variable}
