% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{condemnd}
\alias{condemnd}
\title{R free from conflicting demands}
\source{
General Social Survey https://gss.norc.org
}
\description{
condemnd
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
H. I am free from the conflicting demands that other people make of me
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not too true
\item \verb{[4]} not at all true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2803/vshow}{GSS Data Explorer page for \code{condemnd}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 479 \tab 722 \tab 350 \tab 184 \tab 44 \tab 17 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 430 \tab 739 \tab 352 \tab 137 \tab 57 \tab 19 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 273 \tab 539 \tab 215 \tab 109 \tab 23 \tab 28 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 343 \tab 564 \tab 218 \tab 83 \tab 33 \tab 9 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 400 \tab 653 \tab 259 \tab 77 \tab 24 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 523 \tab 866 \tab 376 \tab 135 \tab 34 \tab 21 \tab 34 \tab 3544 \cr
   Total \tab 8345 \tab 2448 \tab 4083 \tab 1770 \tab 725 \tab 215 \tab 129 \tab 34 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkint}},
\code{\link{fringeok}},
\code{\link{haveinfo}},
\code{\link{hlpequip}},
\code{\link{jobsecok}},
\code{\link{opdevel}},
\code{\link{promtefr}},
\code{\link{supcares}},
\code{\link{wkfreedm}}
}
\concept{Quality of Working Life}
\keyword{variable}
