% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{condrift}
\alias{condrift}
\title{Sci knowledge: the continents have been moving}
\source{
General Social Survey https://gss.norc.org
}
\description{
condrift
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don’t know or aren’t sure, just tell me so,  and we will skip to the next question. Remember true, false, or don’t know.
H. The continents on which we live have been moving their locations for millions of years and will continue to move in the future. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3341/vshow}{GSS Data Explorer page for \code{condrift}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1464 \tab 163 \tab 235 \tab 2 \tab 4510 \cr
   2008 \tab 518 \tab 1148 \tab 155 \tab 193 \tab 9 \tab 2023 \cr
   2010 \tab 1103 \tab 744 \tab 86 \tab 108 \tab 3 \tab 2044 \cr
   2012 \tab 972 \tab 812 \tab 74 \tab 110 \tab 6 \tab 1974 \cr
   2014 \tab 1299 \tab 997 \tab 130 \tab 110 \tab 2 \tab 2538 \cr
   2016 \tab 1477 \tab 1123 \tab 125 \tab 134 \tab 8 \tab 2867 \cr
   2018 \tab 1173 \tab 923 \tab 140 \tab 107 \tab 5 \tab 2348 \cr
   Total \tab 9188 \tab 7211 \tab 873 \tab 997 \tab 35 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
