% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{conf2f}
\alias{conf2f}
\title{How many people r sees face to face}
\source{
General Social Survey https://gss.norc.org
}
\description{
conf2f
}
\details{
Question About how many of these people do you see face-to-face on a typical weekday?
}
\section{Values}{

\itemize{
\item \code{1} all or almost all of them
\item \code{2} most of them
\item \code{3} about half of them
\item \code{4} some of them
\item \code{5} none or almost none of them
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7264/vshow}{GSS Data Explorer page for \code{conf2f}}.
}

\keyword{variable}
