% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{confedvac}
\alias{confedvac}
\title{Confidence in the federal government x}
\source{
General Social Survey https://gss.norc.org
}
\description{
confedvac
}
\details{
Question Here are some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?

The federal government ensuring the safety of vaccines to protect the public against serious diseases
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7580/vshow}{GSS Data Explorer page for \code{confedvac}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab only some \tab hardly any \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2689 \tab 436 \tab 567 \tab 156 \tab 1 \tab 167 \tab 16 \tab 4032 \cr
   Total \tab 2689 \tab 436 \tab 567 \tab 156 \tab 1 \tab 167 \tab 16 \tab 4032 \cr
}
}

\keyword{variable}
