% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{confedvacy}
\alias{confedvacy}
\title{Confidence in the federal government y}
\source{
General Social Survey https://gss.norc.org
}
\description{
confedvacy
}
\details{
Question Here are some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?

The federal government ensuring the safety of vaccines to protect the public against COVID-19
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} only some
\item \verb{[3]} hardly any
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2021 \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7581/vshow}{GSS Data Explorer page for \code{confedvacy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a great deal \tab only some \tab hardly any \tab don't know \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   2021 \tab 2658 \tab 480 \tab 499 \tab 206 \tab 1 \tab 169 \tab 19 \tab - \tab 4032 \cr
   2022 \tab 2309 \tab 488 \tab 480 \tab 259 \tab - \tab - \tab 8 \tab - \tab 3544 \cr
   Total \tab 4967 \tab 968 \tab 979 \tab 465 \tab 1 \tab 169 \tab 27 \tab 0 \tab 7576 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{abhelp1}},
\code{\link{abhelp2}},
\code{\link{abhelp3}},
\code{\link{abhelp4}},
\code{\link{abmoral}},
\code{\link{advfront}},
\code{\link{biokids}},
\code{\link{blkdsrv}},
\code{\link{blkmblty}},
\code{\link{blktry}},
\code{\link{brv16}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16dadnum}},
\code{\link{brv16filnum}},
\code{\link{brv16frndnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16grand}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16oth}},
\code{\link{brv16othnum}},
\code{\link{brv16par}},
\code{\link{brv16partnum}},
\code{\link{brv16sib}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{brv16sp}},
\code{\link{brv16spnum}},
\code{\link{brv5}},
\code{\link{brv5bronum}},
\code{\link{brv5child}},
\code{\link{brv5chinum}},
\code{\link{brv5cowknum}},
\code{\link{brv5cuznum}},
\code{\link{brv5dadnum}},
\code{\link{brv5daunum}},
\code{\link{brv5filnum}},
\code{\link{brv5frndnum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5grand}},
\code{\link{brv5milnum}},
\code{\link{brv5momnum}},
\code{\link{brv5oth}},
\code{\link{brv5othnum}},
\code{\link{brv5par}},
\code{\link{brv5partnum}},
\code{\link{brv5sib}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}},
\code{\link{brv5sonnum}},
\code{\link{brv5sp}},
\code{\link{brv5spnum}},
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{conpharvacy}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{famgen_next}},
\code{\link{femself}},
\code{\link{firstkidsex}},
\code{\link{fluvax}},
\code{\link{fund_next}},
\code{\link{gestate}},
\code{\link{happy_next}},
\code{\link{hhtype1_next}},
\code{\link{knwsexor}},
\code{\link{malekids}},
\code{\link{mascself}},
\code{\link{nexteligible}},
\code{\link{nextgen}},
\code{\link{nonbinkids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{perla_self}},
\code{\link{petsseq}},
\code{\link{scientbe}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{toofast}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
