% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_53.R
\name{confedy}
\alias{confedy}
\title{Exec branch fed govt:version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
confedy
}
\details{
Question I am going to name some institutions in this country. Some people have complete confidence in the people running these institutions. Suppose these people are at one end of the scale at point number 1. Other people have no confidence at all in the people running these institutions. Suppose these people are at the other end, at point 7. And, of course, other people have opinions somewhere in between at point 2, 3, 4, 5 or 6. Where would you place yourself on this scale for . . .
READ EACH ITEM, CODE ONE FOR EACH. REPEAT THE QUESTION AS NECESSARY.
E. Executive branch of the federal government
}
\section{Values}{

\itemize{
\item \verb{[1]} complete confidence
\item \verb{[7]} no confidence
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/465/vshow}{GSS Data Explorer page for \code{confedy}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab complete confidence \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no confidence \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 35 \tab 68 \tab 68 \tab 107 \tab 70 \tab 62 \tab 59 \tab 9 \tab 6 \tab 1473 \cr
   Total \tab 989 \tab 35 \tab 68 \tab 68 \tab 107 \tab 70 \tab 62 \tab 59 \tab 9 \tab 6 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{conarmyy}},
\code{\link{conjudgy}},
\code{\link{conlaboy}},
\code{\link{conlegiy}},
\code{\link{conmediy}},
\code{\link{conpresy}},
\code{\link{consciy}},
\code{\link{contvy}},
\code{\link{manners}}
}
\concept{Core}
\keyword{variable}
