% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{confinan}
\alias{confinan}
\title{Confid in banks & financial institutions}
\source{
General Social Survey https://gss.norc.org
}
\description{
confinan
}
\details{
Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?
READ EACH ITEM; CODE ONE FOR EACH.
A. Banks and financial institutions
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
   1996 \tab B/C \tab partial \cr
   1998 \tab B/C \tab partial \cr
   2000 \tab B/C \tab partial \cr
   2002 \tab B/C \tab partial \cr
   2004 \tab B/C \tab partial \cr
   2006 \tab B/C \tab partial \cr
   2008 \tab B/C \tab partial \cr
   2010 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab B/C \tab partial \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/448/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 475 \tab 44 \tab 165 \tab 2 \tab 804 \tab - \tab 1490 \cr
   1976 \tab - \tab 590 \tab 36 \tab 149 \tab 7 \tab 717 \tab - \tab 1499 \cr
   1977 \tab - \tab 640 \tab 27 \tab 135 \tab 4 \tab 724 \tab - \tab 1530 \cr
   1978 \tab - \tab 503 \tab 21 \tab 179 \tab 4 \tab 825 \tab - \tab 1532 \cr
   1980 \tab - \tab 463 \tab 40 \tab 223 \tab 5 \tab 737 \tab - \tab 1468 \cr
   1982 \tab - \tab 477 \tab 44 \tab 300 \tab 10 \tab 1029 \tab - \tab 1860 \cr
   1983 \tab - \tab 376 \tab 36 \tab 253 \tab 2 \tab 932 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 310 \tab 23 \tab 105 \tab 13 \tab 538 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 308 \tab 26 \tab 257 \tab 4 \tab 875 \tab - \tab 1470 \cr
   1987 \tab - \tab 496 \tab 42 \tab 253 \tab 3 \tab 1025 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 266 \tab 24 \tab 132 \tab 2 \tab 573 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 193 \tab 35 \tab 199 \tab 3 \tab 605 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 160 \tab 21 \tab 198 \tab - \tab 520 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 125 \tab 21 \tab 344 \tab 3 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 158 \tab 23 \tab 275 \tab 3 \tab 598 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 356 \tab 25 \tab 403 \tab 9 \tab 1218 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 476 \tab 47 \tab 317 \tab 5 \tab 1080 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 490 \tab 43 \tab 306 \tab 4 \tab 1068 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 551 \tab 46 \tab 266 \tab 9 \tab 1024 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 201 \tab 13 \tab 167 \tab - \tab 531 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 244 \tab 10 \tab 124 \tab - \tab 498 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 601 \tab 28 \tab 286 \tab - \tab 1074 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 263 \tab 15 \tab 281 \tab - \tab 806 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 144 \tab 11 \tab 564 \tab 2 \tab 656 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 149 \tab 12 \tab 498 \tab 1 \tab 678 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 240 \tab 10 \tab 544 \tab - \tab 893 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 274 \tab 9 \tab 614 \tab 1 \tab 1058 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 298 \tab 12 \tab 380 \tab 1 \tab 872 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 482 \tab 4 \tab 586 \tab 2 \tab 1592 \tab 6 \tab 4032 \cr
   2022 \tab 1173 \tab 391 \tab 14 \tab 599 \tab - \tab 1363 \tab 4 \tab 3544 \cr
   2024 \tab 1116 \tab 377 \tab 19 \tab 561 \tab - \tab 1226 \tab 10 \tab 3309 \cr
   Total \tab 27396 \tab 11077 \tab 781 \tab 9663 \tab 99 \tab 26663 \tab 20 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confed}},
\code{\link{conlabor}},
\code{\link{conpress}},
\code{\link{sathealt}}

Other Business: 
\code{\link{conbus}}

Other Confidence: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confed}},
\code{\link{conlabor}},
\code{\link{conpress}}

Other Split Ballots: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confed}},
\code{\link{conlabor}},
\code{\link{conpress}},
\code{\link{satfam}},
\code{\link{satfrnd}},
\code{\link{sathealt}}
}
\concept{Business}
\concept{Confidence}
\concept{Core}
\concept{Financial Institutions And Banks}
\concept{Split Ballots}
\keyword{variable}
