% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{conghlp1}
\alias{conghlp1}
\title{How much congregation help out with ill}
\source{
General Social Survey https://gss.norc.org
}
\description{
conghlp1
}
\details{
Question If you were ill, how much would the people in your congregation help you out?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} some
\item \verb{[3]} a little
\item \verb{[4]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2092/vshow}{GSS Data Explorer page for \code{conghlp1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a great deal \tab some \tab a little \tab none \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 1797 \tab 372 \tab 261 \tab 110 \tab 63 \tab 65 \tab 164 \tab - \tab 2832 \cr
   Total \tab 1797 \tab 372 \tab 261 \tab 110 \tab 63 \tab 65 \tab 164 \tab 0 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{beausprt}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{drama}},
\code{\link{effort}},
\code{\link{feelgod}},
\code{\link{fidgety}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godlove}},
\code{\link{godwatch}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{lesspain}},
\code{\link{localart}},
\code{\link{meditate}},
\code{\link{natarts}},
\code{\link{natlart}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{popmusic}},
\code{\link{privpray}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{rellife}},
\code{\link{relpersn}},
\code{\link{sprtprsn}},
\code{\link{stateart}},
\code{\link{uniongod}},
\code{\link{volarts}},
\code{\link{wrthless}}

Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{madenkid}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numbwfrd}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{othsch}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcand}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}

Other Social Networks: 
\code{\link{acqasian}},
\code{\link{acqattnd}},
\code{\link{acqblack}},
\code{\link{acqbrnda}},
\code{\link{acqchild}},
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqdems}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhme}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmno}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqfmwht}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjans}},
\code{\link{acqjose}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqlinda}},
\code{\link{acqmaria}},
\code{\link{acqmark}},
\code{\link{acqmils}},
\code{\link{acqmyrac}},
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhlib}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}},
\code{\link{acqnhwht}},
\code{\link{acqnoatt}},
\code{\link{acqprisn}},
\code{\link{acqrachl}},
\code{\link{acqreps}},
\code{\link{acqshawn}},
\code{\link{acqsocs}},
\code{\link{acqunemp}},
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvalib}},
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhme}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkno}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}},
\code{\link{acqwkwht}},
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{comamind}},
\code{\link{conghlp2}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{fricall}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndknow}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{grpother}},
\code{\link{grpsprts}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{hlpinflu}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{lentto}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{peocntct}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkedto}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}},
\code{\link{trtasian}},
\code{\link{trtattnd}},
\code{\link{trtblack}},
\code{\link{trtbrnda}},
\code{\link{trtchild}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtcops}},
\code{\link{trtdems}},
\code{\link{trtelecs}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtjose}},
\code{\link{trtkaren}},
\code{\link{trtkeith}},
\code{\link{trtkevin}},
\code{\link{trtlaws}},
\code{\link{trtlib}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtmils}},
\code{\link{trtmyrac}},
\code{\link{trtnoatt}},
\code{\link{trtprisn}},
\code{\link{trtrachl}},
\code{\link{trtreps}},
\code{\link{trtshawn}},
\code{\link{trtsocs}},
\code{\link{trtunemp}},
\code{\link{trtwhite}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{youinflu}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Religion}
\concept{Religion and Culture}
\concept{Social Networks}
\keyword{variable}
