% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{conghlp2}
\alias{conghlp2}
\title{How much congregation help out with prob}
\source{
General Social Survey https://gss.norc.org
}
\description{
conghlp2
}
\details{
Question If you had a problem or were faced with a difficult situation, how much comfort would the people in your congregation be willing to give you?
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} some
\item \code{3} a little
\item \code{4} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2093/vshow}{GSS Data Explorer page for \code{conghlp2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab some \tab a little \tab none \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1797 \tab 435 \tab 240 \tab 82 \tab 48 \tab 64 \tab 166 \tab 2832 \cr
   Total \tab 1797 \tab 435 \tab 240 \tab 82 \tab 48 \tab 64 \tab 166 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{conghlp1}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{meditate}},
\code{\link{privpray}}
}
\concept{Religion and Culture}
\keyword{variable}
