% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{conghrm2}
\alias{conghrm2}
\title{How often congregation critical of things r do}
\source{
General Social Survey https://gss.norc.org
}
\description{
conghrm2
}
\details{
Question How often are the people in your congregation critical of you and the things you do . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} fairly often
\item \verb{[3]} once in a while
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2095/vshow}{GSS Data Explorer page for \code{conghrm2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab fairly often \tab once in a while \tab never \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1797 \tab 23 \tab 27 \tab 164 \tab 649 \tab 8 \tab 164 \tab 2832 \cr
   Total \tab 1797 \tab 23 \tab 27 \tab 164 \tab 649 \tab 8 \tab 164 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{meditate}},
\code{\link{privpray}}
}
\concept{Religion and Culture}
\keyword{variable}
