% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{conimm}
\alias{conimm}
\title{Conflict between natives and immigrants}
\source{
General Social Survey https://gss.norc.org
}
\description{
conimm
}
\details{
Question People born in America and people from other countries who have come to live in America?
}
\section{Values}{

\itemize{
\item \code{1} very strong conflicts
\item \code{2} strong conflicts
\item \code{3} not very strong conflicts
\item \code{4} there are no conflicts
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7511/vshow}{GSS Data Explorer page for \code{conimm}}.
}

\keyword{variable}
