% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{conjobs}
\alias{conjobs}
\title{Employed and unemployed in conflict}
\source{
General Social Survey https://gss.norc.org
}
\description{
conjobs
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
C. The unemployed and people with jobs?
}
\section{Values}{

\itemize{
\item \code{1} very strong conflict
\item \code{2} strong conflict
\item \code{3} not strong conflict
\item \code{4} no conflict
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4308/vshow}{GSS Data Explorer page for \code{conjobs}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very strong conflict \tab strong conflict \tab not strong conflict \tab no conflict \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 164 \tab 579 \tab 588 \tab 111 \tab 65 \tab 57 \tab 1819 \cr
   Total \tab 255 \tab 164 \tab 579 \tab 588 \tab 111 \tab 65 \tab 57 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
