% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{conlabory}
\alias{conlabory}
\title{Confidence in labor unions}
\source{
General Social Survey https://gss.norc.org
}
\description{
conlabory
}
\details{
Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?
READ EACH ITEM; CODE ONE FOR EACH.
F. Labor unions
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} only some
\item \verb{[3]} hardly any
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8113/vshow}{GSS Data Explorer page for \code{conlabory}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab a great deal \tab only some \tab hardly any \tab don't know \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 176 \tab 561 \tab 281 \tab 53 \tab 2227 \tab 11 \tab 3309 \cr
   Total \tab 0 \tab 176 \tab 561 \tab 281 \tab 53 \tab 2227 \tab 11 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{marsamey}},
\code{\link{norelgsp16}},
\code{\link{pawrkgrw}},
\code{\link{racdif5}}
}
\concept{Core}
\keyword{variable}
