% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_53.R
\name{conlegiy}
\alias{conlegiy}
\title{Congress:version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
conlegiy
}
\details{
Question I am going to name some institutions in this country. Some people have complete confidence in the people running these institutions. Suppose these people are at one end of the scale at point number 1. Other people have no confidence at all in the people running these institutions. Suppose these people are at the other end, at point 7. And, of course, other people have opinions somewhere in between at point 2, 3, 4, 5 or 6. Where would you place yourself on this scale for . . .
READ EACH ITEM, CODE ONE FOR EACH. REPEAT THE QUESTION AS NECESSARY.
L. Congress
}
\section{Values}{

\itemize{
\item \code{1} complete confidence
\item \code{7} no confidence
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/472/vshow}{GSS Data Explorer page for \code{conlegiy}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab complete confidence \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no confidence \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 20 \tab 53 \tab 91 \tab 129 \tab 82 \tab 55 \tab 34 \tab 10 \tab 10 \tab 1473 \cr
   Total \tab 989 \tab 20 \tab 53 \tab 91 \tab 129 \tab 82 \tab 55 \tab 34 \tab 10 \tab 10 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{conarmyy}},
\code{\link{confedy}},
\code{\link{conjudgy}},
\code{\link{conlaboy}},
\code{\link{conmediy}},
\code{\link{conpresy}},
\code{\link{consciy}},
\code{\link{contvy}},
\code{\link{manners}}
}
\concept{Core}
\keyword{variable}
