% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_326.R
\name{conoffcl}
\alias{conoffcl}
\title{Have ever contacted an elected official}
\source{
General Social Survey https://gss.norc.org
}
\description{
conoffcl
}
\details{
Question Over the past 5 years have you done any of the following to express your opinion about an issue or your support for a cause?
D. Contact an elected official by phone, letter, or e-mail
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2913/vshow}{GSS Data Explorer page for \code{conoffcl}}.
}

\seealso{
Other Adult Transitions: 
\code{\link{boycott}},
\code{\link{getmar}},
\code{\link{getmar1}},
\code{\link{havchld}},
\code{\link{havchld1}},
\code{\link{protest}},
\code{\link{signpet}},
\code{\link{supfam}},
\code{\link{supfam1}}

Other Demonstrations: 
\code{\link{boycott}},
\code{\link{protest}},
\code{\link{signpet}}
}
\concept{Adult Transitions}
\concept{Demonstrations}
\keyword{variable}
