% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{conorelgsp}
\alias{conorelgsp}
\title{Partner no religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
conorelgsp
}
\details{
Question Does your (#_spfill1) consider themself to be atheist, agnostic, something else, or nothing in particular?
}
\section{Values}{

\itemize{
\item \verb{[1]} atheist
\item \verb{[2]} agnostic
\item \verb{[3]} something else
\item \verb{[4]} nothing in particular
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8239/vshow}{GSS Data Explorer page for \code{conorelgsp}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab atheist \tab agnostic \tab something else \tab nothing in particular \tab iap \tab Total \cr
   2024 \tab - \tab 20 \tab 13 \tab 7 \tab 72 \tab 3197 \tab 3309 \cr
   Total \tab 0 \tab 20 \tab 13 \tab 7 \tab 72 \tab 3197 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{family16sex}},
\code{\link{femsee1}},
\code{\link{mascsee1}},
\code{\link{mighist}},
\code{\link{racdif5x}},
\code{\link{sexnow2}}

Other Religion: 
\code{\link{reltrad}},
\code{\link{reltrad16}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
