% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{conpharvac}
\alias{conpharvac}
\title{Confidence in pharmaceutical companies x}
\source{
General Social Survey https://gss.norc.org
}
\description{
conpharvac
}
\details{
Question Here are some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?

Pharmaceutical companies working to create safe vaccines to protect the public against serious diseases
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} only some
\item \verb{[3]} hardly any
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7578/vshow}{GSS Data Explorer page for \code{conpharvac}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab only some \tab hardly any \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2689 \tab 541 \tab 519 \tab 102 \tab 2 \tab 167 \tab 12 \tab 4032 \cr
   Total \tab 2689 \tab 541 \tab 519 \tab 102 \tab 2 \tab 167 \tab 12 \tab 4032 \cr
}
}

\keyword{variable}
