% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{conrinc}
\alias{conrinc}
\title{Respondent income in constant dollars}
\source{
General Social Survey https://gss.norc.org
}
\description{
conrinc
}
\details{
Question Inflation-adjusted personal income.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Income: 
\code{\link{coninc}},
\code{\link{incdef}},
\code{\link{minfour}},
\code{\link{povline}},
\code{\link{realinc}},
\code{\link{realrinc}}
}
\concept{Income}
\keyword{variable}
