% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{conschls}
\alias{conschls}
\title{Confidence in schools and educ system}
\source{
General Social Survey https://gss.norc.org
}
\description{
conschls
}
\details{
Question How much confidence do you have in. . .
F. Schools and the educational system
}
\section{Values}{

\itemize{
\item \code{1} complete confidence
\item \code{2} a great deal of confidence
\item \code{3} some confidence
\item \code{4} very little confidence
\item \code{5} no confidence at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4644/vshow}{GSS Data Explorer page for \code{conschls}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab complete confidence \tab a great deal of confidence \tab some confidence \tab very little confidence \tab no confidence at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 119 \tab 362 \tab 574 \tab 195 \tab 46 \tab 32 \tab 31 \tab 1517 \cr
   1998 \tab 1548 \tab 65 \tab 351 \tab 581 \tab 180 \tab 48 \tab 23 \tab 36 \tab 2832 \cr
   2008 \tab 658 \tab 85 \tab 343 \tab 672 \tab 208 \tab 46 \tab 10 \tab 1 \tab 2023 \cr
   2018 \tab 1173 \tab 54 \tab 304 \tab 602 \tab 172 \tab 31 \tab 6 \tab 6 \tab 2348 \cr
   Total \tab 3537 \tab 323 \tab 1360 \tab 2429 \tab 755 \tab 171 \tab 71 \tab 74 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
