% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_52.R
\name{consci}
\alias{consci}
\title{Confidence in scientific community}
\source{
General Social Survey https://gss.norc.org
}
\description{
consci
}
\details{
Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?
READ EACH ITEM; CODE ONE FOR EACH.
K. Scientific Community
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/458/vshow}{GSS Data Explorer page for \code{consci}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab only some \tab hardly any \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1973 \tab - \tab 552 \tab 704 \tab 97 \tab 142 \tab 9 \tab - \tab 1504 \cr
   1974 \tab - \tab 667 \tab 558 \tab 99 \tab 157 \tab 3 \tab - \tab 1484 \cr
   1975 \tab - \tab 560 \tab 672 \tab 96 \tab 159 \tab 3 \tab - \tab 1490 \cr
   1976 \tab - \tab 638 \tab 564 \tab 112 \tab 172 \tab 13 \tab - \tab 1499 \cr
   1977 \tab - \tab 624 \tab 696 \tab 83 \tab 119 \tab 8 \tab - \tab 1530 \cr
   1978 \tab - \tab 553 \tab 737 \tab 111 \tab 126 \tab 5 \tab - \tab 1532 \cr
   1980 \tab - \tab 607 \tab 622 \tab 93 \tab 140 \tab 6 \tab - \tab 1468 \cr
   1982 \tab - \tab 645 \tab 886 \tab 133 \tab 179 \tab 17 \tab - \tab 1860 \cr
   1983 \tab - \tab 662 \tab 745 \tab 85 \tab 101 \tab 6 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 436 \tab 430 \tab 54 \tab 56 \tab 13 \tab - \tab 1473 \cr
   1986 \tab - \tab 569 \tab 697 \tab 113 \tab 84 \tab 7 \tab - \tab 1470 \cr
   1987 \tab - \tab 754 \tab 793 \tab 130 \tab 135 \tab 7 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 385 \tab 473 \tab 56 \tab 78 \tab 5 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 414 \tab 460 \tab 63 \tab 93 \tab 5 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 335 \tab 425 \tab 59 \tab 79 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 411 \tab 467 \tab 63 \tab 73 \tab 3 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 395 \tab 496 \tab 71 \tab 90 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 769 \tab 986 \tab 141 \tab 105 \tab 10 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 752 \tab 867 \tab 144 \tab 155 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 756 \tab 855 \tab 149 \tab 145 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 779 \tab 819 \tab 141 \tab 149 \tab 8 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 336 \tab 440 \tab 81 \tab 55 \tab - \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 359 \tab 434 \tab 52 \tab 30 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 807 \tab 943 \tab 140 \tab 98 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 524 \tab 701 \tab 76 \tab 63 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 544 \tab 683 \tab 79 \tab 65 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 529 \tab 643 \tab 92 \tab 73 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 680 \tab 811 \tab 132 \tab 59 \tab 5 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 791 \tab 976 \tab 117 \tab 69 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 681 \tab 725 \tab 100 \tab 55 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 1337 \tab 1141 \tab 176 \tab 4 \tab - \tab 14 \tab 4032 \cr
   2022 \tab 1173 \tab 930 \tab 1088 \tab 296 \tab 51 \tab 3 \tab 3 \tab 3544 \cr
   Total \tab 20145 \tab 19781 \tab 22537 \tab 3434 \tab 3159 \tab 170 \tab 17 \tab 69243 \cr
}
}

\seealso{
Other Core: 
\code{\link{conarmy}},
\code{\link{conbusy}},
\code{\link{conclery}},
\code{\link{coneducy}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conlegis}},
\code{\link{conmedic}},
\code{\link{contv}}
}
\concept{Core}
\keyword{variable}
