% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{conscrp}
\alias{conscrp}
\title{Importance of military service}
\source{
General Social Survey https://gss.norc.org
}
\description{
conscrp
}
\details{
Question On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:... To be willing to serve in the military at a time of need.
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8173/vshow}{GSS Data Explorer page for \code{conscrp}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 117 \tab 70 \tab 86 \tab 212 \tab 235 \tab 207 \tab 590 \tab 106 \tab 1663 \tab 2 \tab 21 \tab 3309 \cr
   Total \tab 0 \tab 117 \tab 70 \tab 86 \tab 212 \tab 235 \tab 207 \tab 590 \tab 106 \tab 1663 \tab 2 \tab 21 \tab 3309 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{bizmoney}},
\code{\link{conpow}},
\code{\link{experts}},
\code{\link{fulldem}},
\code{\link{immwlfare}},
\code{\link{incgap1}}
}
\concept{ISSP National Identity}
\keyword{variable}
