% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{consent}
\alias{consent}
\title{Participation/recording consent}
\source{
General Social Survey https://gss.norc.org
}
\description{
consent
}
\details{
Question PARTICIPATION/RECORDING CONSENT
}
\section{Values}{

\itemize{
\item \verb{[1]} r consents to recording interview
\item \verb{[2]} r consents to participate in the survey, but does not want to be recorded
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5569/vshow}{GSS Data Explorer page for \code{consent}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab r consents to recording interview \tab r consents to participate in the survey, but does not want to be recorded \tab no answer \tab Total \cr
   2008 \tab - \tab 1697 \tab 324 \tab 2 \tab 2023 \cr
   2010 \tab - \tab 1687 \tab 357 \tab - \tab 2044 \cr
   2012 \tab - \tab 1730 \tab 244 \tab - \tab 1974 \cr
   2014 \tab - \tab 2075 \tab 463 \tab - \tab 2538 \cr
   2016 \tab - \tab 2600 \tab 267 \tab - \tab 2867 \cr
   2018 \tab - \tab 2182 \tab 166 \tab - \tab 2348 \cr
   2021 \tab 2322 \tab 1707 \tab 3 \tab - \tab 4032 \cr
   2022 \tab 1636 \tab 1724 \tab 184 \tab - \tab 3544 \cr
   Total \tab 3958 \tab 15402 \tab 2008 \tab 2 \tab 21370 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{devtype}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{huclean}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{intid}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{intyrs_0418}},
\code{\link{lngthinv}},
\code{\link{mode}()},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse1}},
\code{\link{whoelse2}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}

Other Interview: 
\code{\link{comprend}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{id}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{inthisp}},
\code{\link{intid}},
\code{\link{intrace1}},
\code{\link{intrace2}},
\code{\link{intrace3}},
\code{\link{intrwght}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{intyrs_0418}},
\code{\link{lngthinv}},
\code{\link{mode}()},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{sciintro}},
\code{\link{scimode}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{year}()}
}
\concept{Interview}
\concept{Paradata}
\keyword{variable}
