% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{contmslm}
\alias{contmslm}
\title{Have muslims positive contribution to country}
\source{
General Social Survey https://gss.norc.org
}
\description{
contmslm
}
\details{
Question Since the beginning of our country, people of many different races, nationalities, and religions have come here. and settled. As I name some of these groups please tell me if the group has made one of the most important positive contributions to this country, an important contribution, some contribution, or little positive contribution to this community.
L. Muslims?
}
\section{Values}{

\itemize{
\item \verb{[1]} most important contribution
\item \verb{[2]} important contribution
\item \verb{[3]} some contribution
\item \verb{[4]} little positive contribution
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2282/vshow}{GSS Data Explorer page for \code{contmslm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most important contribution \tab important contribution \tab some contribution \tab little positive contribution \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 32 \tab 202 \tab 462 \tab 406 \tab 286 \tab 10 \tab 2817 \cr
   Total \tab 1419 \tab 32 \tab 202 \tab 462 \tab 406 \tab 286 \tab 10 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
