% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{contpoor}
\alias{contpoor}
\title{How often r has contact with poor people}
\source{
General Social Survey https://gss.norc.org
}
\description{
contpoor
}
\details{
Question How often do you have any contact with people who are a lot poorer than you when you are out and about? This might be in the street, on public transport, in shops, in your neighborhood, or at your workplace.
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than once a month
\item \verb{[3]} once a month
\item \verb{[4]} several times a month
\item \verb{[5]} once a week
\item \verb{[6]} several times a week
\item \verb{[7]} every day
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7516/vshow}{GSS Data Explorer page for \code{contpoor}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab never \tab less than once a month \tab once a month \tab several times a month \tab once a week \tab several times a week \tab every day \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 80 \tab 240 \tab 151 \tab 378 \tab 128 \tab 377 \tab 350 \tab 112 \tab 174 \tab 13 \tab 4032 \cr
   Total \tab 2029 \tab 80 \tab 240 \tab 151 \tab 378 \tab 128 \tab 377 \tab 350 \tab 112 \tab 174 \tab 13 \tab 4032 \cr
}
}

\keyword{variable}
