% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_252.R
\name{contpr}
\alias{contpr}
\title{Have puerto ricans positive contribution to country}
\source{
General Social Survey https://gss.norc.org
}
\description{
contpr
}
\details{
Question Since the beginning of our country, people of many different races, nationalities, and religions have come here. and settled. As I name some of these groups please tell me if the group has made one of the most important positive contributions to this country, an important contribution, some contribution, or little positive contribution to this community.
J. Puerto Ricans?
}
\section{Values}{

\itemize{
\item \code{1} most important contribution
\item \code{2} important contribution
\item \code{3} some contribution
\item \code{4} little positive contribution
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2280/vshow}{GSS Data Explorer page for \code{contpr}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most important contribution \tab important contribution \tab some contribution \tab little positive contribution \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 35 \tab 224 \tab 563 \tab 369 \tab 196 \tab 11 \tab 2817 \cr
   Total \tab 1419 \tab 35 \tab 224 \tab 563 \tab 369 \tab 196 \tab 11 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contmex}},
\code{\link{contvn}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
