% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{contrich}
\alias{contrich}
\title{How often r has contact with rich people}
\source{
General Social Survey https://gss.norc.org
}
\description{
contrich
}
\details{
Question How often do you have any contact with people who are a lot richer than you when you are out and about? This might be in the street, on public transport, in shops, in your neighborhood, or at your workplace.
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than once a month
\item \verb{[3]} once a month
\item \verb{[4]} several times a month
\item \verb{[5]} once a week
\item \verb{[6]} several times a week
\item \verb{[7]} every day
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7517/vshow}{GSS Data Explorer page for \code{contrich}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab never \tab less than once a month \tab once a month \tab several times a month \tab once a week \tab several times a week \tab every day \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 158 \tab 349 \tab 188 \tab 399 \tab 110 \tab 294 \tab 208 \tab 104 \tab 175 \tab 18 \tab 4032 \cr
   Total \tab 2029 \tab 158 \tab 349 \tab 188 \tab 399 \tab 110 \tab 294 \tab 208 \tab 104 \tab 175 \tab 18 \tab 4032 \cr
}
}

\keyword{variable}
