% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_54.R
\name{control}
\alias{control}
\title{Self-control}
\source{
General Social Survey https://gss.norc.org
}
\description{
control
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
6. That he has self-control
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 most desirable
\item \verb{[2]} 3 most desirable
\item \verb{[3]} not mentioned
\item \verb{[4]} 3 least desirable
\item \verb{[5]} 1 least desirable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1975, 1976, 1978, 1980, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/479/vshow}{GSS Data Explorer page for \code{control}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 46 \tab 221 \tab 1030 \tab 174 \tab 29 \tab 4 \tab - \tab 1504 \cr
   1975 \tab - \tab 53 \tab 228 \tab 1052 \tab 124 \tab 23 \tab 10 \tab - \tab 1490 \cr
   1976 \tab - \tab 40 \tab 217 \tab 1046 \tab 161 \tab 26 \tab 9 \tab - \tab 1499 \cr
   1978 \tab - \tab 45 \tab 240 \tab 1084 \tab 125 \tab 26 \tab 10 \tab 2 \tab 1532 \cr
   1980 \tab 975 \tab 13 \tab 57 \tab 354 \tab 57 \tab 8 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 42 \tab 196 \tab 1183 \tab 132 \tab 26 \tab 20 \tab - \tab 1599 \cr
   Total \tab 975 \tab 239 \tab 1159 \tab 5749 \tab 773 \tab 138 \tab 56 \tab 3 \tab 9092 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicable}},
\code{\link{clean}},
\code{\link{consider}},
\code{\link{honest}},
\code{\link{judgment}},
\code{\link{obeys}},
\code{\link{responsi}},
\code{\link{role}},
\code{\link{success}}
}
\concept{Core}
\keyword{variable}
