% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{conunion}
\alias{conunion}
\title{Workers and management in conflict?}
\source{
General Social Survey https://gss.norc.org
}
\description{
conunion
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
D. Management and workers?
}
\section{Values}{

\itemize{
\item \verb{[1]} very strong conflicts
\item \verb{[2]} strong conflicts
\item \verb{[3]} not very strong conflicts
\item \verb{[4]} there are no conflicts
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4309/vshow}{GSS Data Explorer page for \code{conunion}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very strong conflicts \tab strong conflicts \tab not very strong conflicts \tab there are no conflicts \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 167 \tab 655 \tab 561 \tab 53 \tab 74 \tab 54 \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 142 \tab 463 \tab 463 \tab 31 \tab 59 \tab 114 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 140 \tab 721 \tab 780 \tab 46 \tab 142 \tab 159 \tab 15 \tab 4032 \cr
   Total \tab 3829 \tab 449 \tab 1839 \tab 1804 \tab 130 \tab 275 \tab 327 \tab 15 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
