% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{conurban}
\alias{conurban}
\title{Farmer's and city people in conflict?}
\source{
General Social Survey https://gss.norc.org
}
\description{
conurban
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
E. Farmers and city people?
}
\section{Values}{

\itemize{
\item \verb{[1]} very strong conflict
\item \verb{[2]} strong conflict
\item \verb{[3]} not strong conflict
\item \verb{[4]} no conflict
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4310/vshow}{GSS Data Explorer page for \code{conurban}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very strong conflict \tab strong conflict \tab not strong conflict \tab no conflict \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 135 \tab 436 \tab 669 \tab 185 \tab 93 \tab 46 \tab 1819 \cr
   Total \tab 255 \tab 135 \tab 436 \tab 669 \tab 185 \tab 93 \tab 46 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
