% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{convictd}
\alias{convictd}
\title{Convicted of crime ever}
\source{
General Social Survey https://gss.norc.org
}
\description{
convictd
}
\details{
Question B. Not counting minor traffic offenses, have you ever been convicted of a crime?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/686/vshow}{GSS Data Explorer page for \code{convictd}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2012 \tab 220 \tab 210 \tab 1538 \tab 1 \tab 5 \tab 1974 \cr
   Total \tab 220 \tab 210 \tab 1538 \tab 1 \tab 5 \tab 1974 \cr
}
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}},
\code{\link{tvhours}}
}
\concept{Core}
\keyword{variable}
