% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{conwkday}
\alias{conwkday}
\title{Number of people r contacts with on a typical weekday}
\source{
General Social Survey https://gss.norc.org
}
\description{
conwkday
}
\details{
Question Please indicate about how many people do you have contact with on a typical weekday irrespective of whether you know them or not. Include anyone you chat with, talk to, or text, either face-to-face, by phone, internet or any other communication device.
}
\section{Values}{

\itemize{
\item \verb{[1]} 0-4 people
\item \verb{[2]} 5-9 people
\item \verb{[3]} 10-19 people
\item \verb{[4]} 20-49 people
\item \verb{[5]} 50-99 people
\item \verb{[6]} 100 or more people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7263/vshow}{GSS Data Explorer page for \code{conwkday}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 0-4 people \tab 5-9 people \tab 10-19 people \tab 20-49 people \tab 50-99 people \tab 100 or more people \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 226 \tab 313 \tab 268 \tab 224 \tab 67 \tab 70 \tab 4 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 226 \tab 313 \tab 268 \tab 224 \tab 67 \tab 70 \tab 4 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
