% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{cooking1}
\alias{cooking1}
\title{Who in household prepares the meals}
\source{
General Social Survey https://gss.norc.org
}
\description{
cooking1
}
\details{
Question In your household who does the following things...?
F. Prepares the meals?
}
\section{Values}{

\itemize{
\item \verb{[1]} always me
\item \verb{[2]} usually me
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually my spouse/partner
\item \verb{[5]} always my spouse/partner
\item \verb{[6]} is done by a third person
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4434/vshow}{GSS Data Explorer page for \code{cooking1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab always me \tab usually me \tab about equal or both together \tab usually my spouse/partner \tab always my spouse/partner \tab is done by a third person \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 2058 \tab 155 \tab 152 \tab 207 \tab 157 \tab 9 \tab 3 \tab 2 \tab 22 \tab - \tab 2765 \cr
   2012 \tab 1274 \tab 160 \tab 133 \tab 192 \tab 123 \tab 75 \tab 13 \tab 1 \tab 3 \tab - \tab 1974 \cr
   2022 \tab 3024 \tab 86 \tab 110 \tab 152 \tab 115 \tab 46 \tab 3 \tab 5 \tab 1 \tab 2 \tab 3544 \cr
   Total \tab 6356 \tab 401 \tab 395 \tab 551 \tab 395 \tab 130 \tab 19 \tab 8 \tab 26 \tab 2 \tab 8283 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{deckids}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{spborn}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
