% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{coop2}
\alias{coop2}
\title{R's attitude toward interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
coop2
}
\details{
Question In general, what was the respondent's attitude toward the interview?
}
\section{Values}{

\itemize{
\item \verb{[1]} friendly and eager
\item \verb{[2]} cooperative
\item \verb{[3]} indifferent
\item \verb{[4]} hostile
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/694/vshow}{GSS Data Explorer page for \code{coop2}}.

Counts by year:\tabular{llllllll}{
   year \tab friendly and eager \tab cooperative \tab indifferent \tab hostile \tab no answer \tab iap \tab Total \cr
   1972 \tab 953 \tab 600 \tab 43 \tab 4 \tab 13 \tab - \tab 1613 \cr
   Total \tab 953 \tab 600 \tab 43 \tab 4 \tab 13 \tab 0 \tab 1613 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abspno}},
\code{\link{comprend}},
\code{\link{coop}},
\code{\link{form}},
\code{\link{phone}},
\code{\link{radiohrs}}
}
\concept{Core}
\keyword{variable}
