% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{cope2}
\alias{cope2}
\title{Work together with god as partner's}
\source{
General Social Survey https://gss.norc.org
}
\description{
cope2
}
\details{
Question Think about how you try to understand and deal with major problems in your life. To what extent is each of the following involved in the way you cope:
B. I work together with God as partners.
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} quite a bit
\item \code{3} somewhat
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2097/vshow}{GSS Data Explorer page for \code{cope2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab quite a bit \tab somewhat \tab not at all \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 312 \tab 316 \tab 502 \tab 267 \tab 29 \tab 19 \tab 2832 \cr
   Total \tab 1387 \tab 312 \tab 316 \tab 502 \tab 267 \tab 29 \tab 19 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{meditate}},
\code{\link{privpray}}
}
\concept{Religion and Culture}
\keyword{variable}
