% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{cope5}
\alias{cope5}
\title{Wonder whether god has abandoned me}
\source{
General Social Survey https://gss.norc.org
}
\description{
cope5
}
\details{
Question Think about how you try to understand and deal with major problems in your life. To what extent is each of the following involved in the way you cope:
E. I wonder whether God has abandoned me.
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} quite a bit
\item \verb{[3]} somewhat
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2100/vshow}{GSS Data Explorer page for \code{cope5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab quite a bit \tab somewhat \tab not at all \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 17 \tab 24 \tab 134 \tab 1232 \tab 15 \tab 23 \tab 2832 \cr
   Total \tab 1387 \tab 17 \tab 24 \tab 134 \tab 1232 \tab 15 \tab 23 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{relcmfrt}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
