% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{copres10}
\alias{copres10}
\title{Partner's occupational prestige score (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
copres10
}
\details{
Question COPRES10 is based on the 2010 Census occupation classification. This standard prestige score is a simple mean value of ratings for each occupation category, converted to a scale of 0 (bottom) to 100 (top).
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\concept{Core}
\keyword{variable}
