% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{corrupt2}
\alias{corrupt2}
\title{Opinion of corruption by government administrators}
\source{
General Social Survey https://gss.norc.org
}
\description{
corrupt2
}
\details{
Question And in your opinion, about how many government administrators in America are involved in corruption?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost none
\item \verb{[2]} a few
\item \verb{[3]} some
\item \verb{[4]} quite a lot
\item \verb{[5]} almost all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4124/vshow}{GSS Data Explorer page for \code{corrupt2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost none \tab a few \tab some \tab quite a lot \tab almost all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 55 \tab 339 \tab 581 \tab 378 \tab 109 \tab 52 \tab 4 \tab 4510 \cr
   2016 \tab 1477 \tab 29 \tab 318 \tab 529 \tab 336 \tab 128 \tab 36 \tab 14 \tab 2867 \cr
   Total \tab 4469 \tab 84 \tab 657 \tab 1110 \tab 714 \tab 237 \tab 88 \tab 18 \tab 7377 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bribe}},
\code{\link{corrupt1}},
\code{\link{knowpols}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
