% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{cosector}
\alias{cosector}
\title{Type of college partner attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
cosector
}
\details{
Question In what year did your partner receive their (COLLEGE NAME) degree?
I. Sector of Institution?
}
\section{Values}{

\itemize{
\item \verb{[0]} administrative unit
\item \verb{[1]} public, 4-year or above
\item \verb{[2]} private not-for-profit, 4-year or above
\item \verb{[3]} private for-profit, 4-year or above
\item \verb{[4]} public, 2-year
\item \verb{[5]} private not-for-profit, 2-year or less
\item \verb{[6]} private for-profit, 2-year
\item \verb{[7]} public, less than 2-year
\item \verb{[9]} private for-profit, less than 2-year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6163/vshow}{GSS Data Explorer page for \code{cosector}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab public, 4-year or above \tab private not-for-profit, 4-year or above \tab public, 2-year \tab private for-profit, 4-year or above \tab private for-profit, 2-year \tab skipped on web \tab Total \cr
   2012 \tab 1939 \tab 23 \tab 11 \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2506 \tab 11 \tab 13 \tab 4 \tab 3 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 2806 \tab 29 \tab 21 \tab 8 \tab 2 \tab - \tab 1 \tab 2867 \cr
   Total \tab 7251 \tab 63 \tab 45 \tab 13 \tab 5 \tab 1 \tab 1 \tab 7379 \cr
}
}

\seealso{
Other Core: 
\code{\link{codipged}},
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\keyword{variable}
