% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{cosei10inc}
\alias{cosei10inc}
\title{Percentage of $45k+ earners in coocc10 based on acs 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
cosei10inc
}
\details{
Question Percentage of those who earned $45,000 or more among those who had the same occupation as in COOCC10 in American Community Survey between 2010 and 2012 (25-64 years old only).This percentage is not from the GSS but it is used to calculate COSEI10.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} i don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{copres105plus}}
}
\concept{Occupation}
\keyword{variable}
