% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_430.R
\name{counsel2}
\alias{counsel2}
\title{Leader2 was r's counselor}
\source{
General Social Survey https://gss.norc.org
}
\description{
counsel2
}
\details{
Question B. Was this leader someone to whom you went for counseling?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3943/vshow}{GSS Data Explorer page for \code{counsel2}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 2017 \tab 1 \tab 5 \tab 2023 \cr
   Total \tab 2017 \tab 1 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{counsel1}},
\code{\link{counsel3}},
\code{\link{numclerg}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
