% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_430.R
\name{counsel3}
\alias{counsel3}
\title{Leader3 was r's counselor}
\source{
General Social Survey https://gss.norc.org
}
\description{
counsel3
}
\details{
Question C. Was this leader someone to whom you went for counseling?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3944/vshow}{GSS Data Explorer page for \code{counsel3}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab no \tab Total \cr
   2008 \tab 2021 \tab 2 \tab 2023 \cr
   Total \tab 2021 \tab 2 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{numclerg}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
