% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_148.R
\name{country}
\alias{country}
\title{Like or dislike country western music}
\source{
General Social Survey https://gss.norc.org
}
\description{
country
}
\details{
Question I'm going to read you a list of some types of music. Can you tell me which of the statements on this card comes closest to your feeling about each type of music (HAND CARD CA TO RESPONDENT.) Let's start with big band music. Do you like it very much, like it, have mixed feelings, dislike it, dislike it very much, or is this a type of music that you don't know much about?
C. Country/Western
}
\section{Values}{

\itemize{
\item \code{1} like very much
\item \code{2} like it
\item \code{3} mixed feelings
\item \code{4} dislike it
\item \code{5} dislike very much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1389/vshow}{GSS Data Explorer page for \code{country}}.
}

\seealso{
Other Culture: 
\code{\link{blues}},
\code{\link{classicl}},
\code{\link{folk}},
\code{\link{gospel}},
\code{\link{jazz}},
\code{\link{latin}},
\code{\link{moodeasy}},
\code{\link{musicals}},
\code{\link{newage}}

Other Music: 
\code{\link{blues}},
\code{\link{classicl}},
\code{\link{folk}},
\code{\link{gospel}},
\code{\link{jazz}},
\code{\link{latin}},
\code{\link{moodeasy}},
\code{\link{musicals}},
\code{\link{newage}}
}
\concept{Culture}
\concept{Music}
\keyword{variable}
