% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{courtsv}
\alias{courtsv}
\title{Courts dealing with criminals (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
courtsv
}
\details{
Question In general, do you think the courts in this area deal too harshly or not harshly enough with criminals?
}
\section{Values}{

\itemize{
\item \verb{[1]} too harshly
\item \verb{[2]} not harshly enough
\item \verb{[3]} about right (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7586/vshow}{GSS Data Explorer page for \code{courtsv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too harshly \tab not harshly enough \tab about right (phone mode only: volunteered) \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 1987 \tab 424 \tab 965 \tab 590 \tab 12 \tab 54 \tab - \tab 4032 \cr
   2022 \tab 2683 \tab 111 \tab 405 \tab 337 \tab 1 \tab 6 \tab 1 \tab 3544 \cr
   Total \tab 4670 \tab 535 \tab 1370 \tab 927 \tab 13 \tab 60 \tab 1 \tab 7576 \cr
}
}

\keyword{variable}
