% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{courtsy}
\alias{courtsy}
\title{Courts dealing with criminals-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
courtsy
}
\details{
Question In general, do you think the courts in this area deal too harshly, or not harshly enough with criminals, or don't you have enough information about the courts to say?
}
\section{Values}{

\itemize{
\item \code{1} too harsh
\item \code{2} not harsh enough
\item \code{3} about right
\item \code{8} cant say
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab - \tab full \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/277/vshow}{GSS Data Explorer page for \code{courtsy}}.
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}}

Other Courts: 
\code{\link{courts}}

Other Crime: 
\code{\link{courts}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}}

Other Split Ballots: 
\code{\link{courts}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}}
}
\concept{Core}
\concept{Courts}
\concept{Crime}
\concept{Split Ballots}
\keyword{variable}
