% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{covemply}
\alias{covemply}
\title{Change in employment over covid pandemic}
\source{
General Social Survey https://gss.norc.org
}
\description{
covemply
}
\details{
Question Considering your work activity before the COVID-19 pandemic and at present, which of the following statements best describes your employment situation?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} r did not have job before pandemic and does not have job now
\item \verb{[2]} r has the same job as before the pandemic
\item \verb{[3]} r lost job due to the pandemic and now has a new job
\item \verb{[4]} r lost job due to the pandemic and could not find a new job
\item \verb{[5]} r did not have job before the pandemic and does have job now
\item \verb{[6]} r changed or left job for reasons unrelated to the pandemic
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7953/vshow}{GSS Data Explorer page for \code{covemply}}.

Counts by year:\tabular{lllllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab Total \cr
   2022 \tab 2392 \tab 294 \tab 456 \tab 67 \tab 37 \tab 41 \tab 257 \tab 3544 \cr
   Total \tab 2392 \tab 294 \tab 456 \tab 67 \tab 37 \tab 41 \tab 257 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covid12}},
\code{\link{fluvax}},
\code{\link{malekids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
