% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{covid12}
\alias{covid12}
\title{Covid vaccine ever}
\source{
General Social Survey https://gss.norc.org
}
\description{
covid12
}
\details{
Question Have you ever received a COVID-19 vaccine? (Regardless of how many doses.)

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7952/vshow}{GSS Data Explorer page for \code{covid12}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 999 \tab 227 \tab 9 \tab 3544 \cr
   Total \tab 2309 \tab 999 \tab 227 \tab 9 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covemply}},
\code{\link{fluvax}},
\code{\link{malekids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
