% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{cowrkers}
\alias{cowrkers}
\title{Quality of rels among workers: r's firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkers
}
\details{
Question In general how would you describe relations in your work place…
B. between co-workers/colleagues?
}
\section{Values}{

\itemize{
\item \verb{[1]} very good
\item \verb{[2]} quite good
\item \verb{[3]} neither good nor bad
\item \verb{[4]} quite bad
\item \verb{[5]} very bad
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4560/vshow}{GSS Data Explorer page for \code{cowrkers}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very good \tab quite good \tab neither good nor bad \tab quite bad \tab very bad \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 239 \tab 422 \tab 159 \tab 16 \tab 2 \tab 16 \tab 44 \tab 1537 \cr
   1998 \tab 1967 \tab 261 \tab 391 \tab 123 \tab 13 \tab 4 \tab 21 \tab 52 \tab 2832 \cr
   2006 \tab 3492 \tab 400 \tab 426 \tab 154 \tab 18 \tab 11 \tab 7 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 333 \tab 405 \tab 151 \tab 27 \tab 6 \tab 9 \tab 9 \tab 2867 \cr
   Total \tab 8025 \tab 1233 \tab 1644 \tab 587 \tab 74 \tab 23 \tab 53 \tab 107 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
