% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{cowrkhlp}
\alias{cowrkhlp}
\title{Coworkers can be relied on when r needs help}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkhlp
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
N. The people I work with can be relied on when I need help
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2809/vshow}{GSS Data Explorer page for \code{cowrkhlp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 957 \tab 643 \tab 106 \tab 55 \tab 14 \tab 21 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 842 \tab 689 \tab 93 \tab 61 \tab 29 \tab 20 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 590 \tab 445 \tab 78 \tab 42 \tab 5 \tab 27 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 646 \tab 463 \tab 79 \tab 45 \tab 8 \tab 9 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 731 \tab 538 \tab 94 \tab 38 \tab 12 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 962 \tab 765 \tab 136 \tab 68 \tab 22 \tab 13 \tab 23 \tab 3544 \cr
   Total \tab 8345 \tab 4728 \tab 3543 \tab 586 \tab 309 \tab 90 \tab 125 \tab 23 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
