% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{cowrkint}
\alias{cowrkint}
\title{Coworkers take a personal interest in r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkint
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
J. The people I work with take a personal interest in me
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not too true
\item \verb{[4]} not at all true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2805/vshow}{GSS Data Explorer page for \code{cowrkint}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 789 \tab 722 \tab 155 \tab 89 \tab 20 \tab 21 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 699 \tab 728 \tab 173 \tab 77 \tab 35 \tab 22 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 469 \tab 508 \tab 110 \tab 54 \tab 18 \tab 28 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 493 \tab 574 \tab 114 \tab 46 \tab 13 \tab 10 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 588 \tab 621 \tab 131 \tab 61 \tab 12 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 734 \tab 862 \tab 205 \tab 117 \tab 28 \tab 18 \tab 25 \tab 3544 \cr
   Total \tab 8345 \tab 3772 \tab 4015 \tab 888 \tab 444 \tab 126 \tab 134 \tab 25 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{fringeok}},
\code{\link{haveinfo}},
\code{\link{hlpequip}},
\code{\link{jobsecok}},
\code{\link{opdevel}},
\code{\link{promtefr}},
\code{\link{supcares}},
\code{\link{wkfreedm}}
}
\concept{Quality of Working Life}
\keyword{variable}
