% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{credam}
\alias{credam}
\title{Americans' credit score has to do with choices or things out of their control}
\source{
General Social Survey https://gss.norc.org
}
\description{
credam
}
\details{
Question Now, thinking about Americans in general, do you think credit scores are more the result of choices people make or things out of their control?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} entirely choices they make
\item \verb{[2]} mostly choices they make
\item \verb{[3]} both choices they make and things out of their control
\item \verb{[4]} mostly things out of their control
\item \verb{[5]} entirely things out of their control
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8263/vshow}{GSS Data Explorer page for \code{credam}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab entirely choices they make \tab mostly choices they make \tab both choices they make and things out of their control \tab mostly things out of their control \tab entirely things out of their control \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 143 \tab 360 \tab 505 \tab 51 \tab 4 \tab 2244 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 143 \tab 360 \tab 505 \tab 51 \tab 4 \tab 2244 \tab 2 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{petsseq}}

Other Income: 
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}}
}
\concept{GSS Next}
\concept{Income}
\keyword{variable}
