% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{credcar}
\alias{credcar}
\title{Use credit report - car insurers}
\source{
General Social Survey https://gss.norc.org
}
\description{
credcar
}
\details{
Question (It’s not only lenders that use credit reports to make decisions. How fair or unfair do you think it is for each of the following to use credit reports?) Car insurers deciding who to insure.
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} very fair
\item \verb{[2]} somewhat fair
\item \verb{[3]} neither fair nor unfair
\item \verb{[4]} somewhat unfair
\item \verb{[5]} very unfair
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8266/vshow}{GSS Data Explorer page for \code{credcar}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab very fair \tab somewhat fair \tab neither fair nor unfair \tab somewhat unfair \tab very unfair \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 128 \tab 280 \tab 176 \tab 267 \tab 212 \tab 2244 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 128 \tab 280 \tab 176 \tab 267 \tab 212 \tab 2244 \tab 2 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{credam}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{petsseq}}

Other Income: 
\code{\link{credam}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}}
}
\concept{GSS Next}
\concept{Income}
\keyword{variable}
