% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{credrep}
\alias{credrep}
\title{R credit report}
\source{
General Social Survey https://gss.norc.org
}
\description{
credrep
}
\details{
Question In general, would you say that your credit record is poor, fair, good, very good, or excellent?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} poor
\item \verb{[2]} fair
\item \verb{[3]} good
\item \verb{[4]} very good
\item \verb{[5]} excellent
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8290/vshow}{GSS Data Explorer page for \code{credrep}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab poor \tab fair \tab good \tab very good \tab excellent \tab can't choose \tab iap \tab Total \cr
   2024 \tab - \tab 97 \tab 141 \tab 172 \tab 161 \tab 465 \tab 29 \tab 2244 \tab 3309 \cr
   Total \tab 0 \tab 97 \tab 141 \tab 172 \tab 161 \tab 465 \tab 29 \tab 2244 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{petsseq}}

Other Income: 
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}}
}
\concept{GSS Next}
\concept{Income}
\keyword{variable}
