% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{crimhold}
\alias{crimhold}
\title{Should police detain a known criminal?}
\source{
General Social Survey https://gss.norc.org
}
\description{
crimhold
}
\details{
Question Suppose the police get an anonymous tip that a man with a long criminal record is planning to break into a warehouse. Please circle a number for each question to show if you think police should be allowed without a court order to ...
D. Detain the man overnight for questioning.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely allowed
\item \verb{[2]} probably allowed
\item \verb{[3]} probably not allowed
\item \verb{[4]} definitely not allowed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3997/vshow}{GSS Data Explorer page for \code{crimhold}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 205 \tab 181 \tab 101 \tab 140 \tab 35 \tab 15 \tab 1534 \cr
   1990 \tab 155 \tab 283 \tab 351 \tab 248 \tab 222 \tab 48 \tab 65 \tab 1372 \cr
   Total \tab 1012 \tab 488 \tab 532 \tab 349 \tab 362 \tab 83 \tab 80 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
