% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{criminfo}
\alias{criminfo}
\title{How much info does r have on crime issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
criminfo
}
\details{
Question How much information do you have about the crime issue? Do you have all of the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \verb{[1]} all info r needs
\item \verb{[2]} most of the info
\item \verb{[3]} some info
\item \verb{[4]} very little info
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/279/vshow}{GSS Data Explorer page for \code{criminfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all info r needs \tab most of the info \tab some info \tab very little info \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 55 \tab 140 \tab 228 \tab 53 \tab 3 \tab 5 \tab 1473 \cr
   Total \tab 989 \tab 55 \tab 140 \tab 228 \tab 53 \tab 3 \tab 5 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
