% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{crimlose}
\alias{crimlose}
\title{People convicted of serious crimes lose citizen rights}
\source{
General Social Survey https://gss.norc.org
}
\description{
crimlose
}
\details{
Question On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
B. That people convicted of serious crimes lose their citizen rights
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4967/vshow}{GSS Data Explorer page for \code{crimlose}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1274 \tab 99 \tab 71 \tab 77 \tab 163 \tab 161 \tab 170 \tab 390 \tab 129 \tab 4 \tab 2538 \cr
   Total \tab 1274 \tab 99 \tab 71 \tab 77 \tab 163 \tab 161 \tab 170 \tab 390 \tab 129 \tab 4 \tab 2538 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
