% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{crimread}
\alias{crimread}
\title{Should police open mail of a criminal?}
\source{
General Social Survey https://gss.norc.org
}
\description{
crimread
}
\details{
Question Suppose the police get an anonymous tip that a man with a long criminal record is planning to break into a warehouse. Please circle a number for each question to show if you think police should be allowed without a court order to ...
C. Open his mail.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3996/vshow}{GSS Data Explorer page for \code{crimread}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 77 \tab 68 \tab 155 \tab 327 \tab 33 \tab 17 \tab 1534 \cr
   1990 \tab 155 \tab 93 \tab 126 \tab 340 \tab 532 \tab 59 \tab 67 \tab 1372 \cr
   Total \tab 1012 \tab 170 \tab 194 \tab 495 \tab 859 \tab 92 \tab 84 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
