% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{crimtail}
\alias{crimtail}
\title{Should police tail a known criminal?}
\source{
General Social Survey https://gss.norc.org
}
\description{
crimtail
}
\details{
Question Suppose the police get an anonymous tip that a man with a long criminal record is planning to break into a warehouse. Please circle a number for each question to show if you think police should be allowed without a court order to ...
A. Keep the man under surveillance.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3994/vshow}{GSS Data Explorer page for \code{crimtail}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}}

Other Civil Liberties: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}}

Other Privacy: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtap}}
}
\concept{Civil Liberties}
\concept{ISSP Role of Gov't}
\concept{Privacy}
\keyword{variable}
