% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{crtrheal}
\alias{crtrheal}
\title{Contrast between creator and healer}
\source{
General Social Survey https://gss.norc.org
}
\description{
crtrheal
}
\details{
Question There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? (HAND CARD AE) The first set of contrasting images shows Mother at 1 on the scale and Father at 7. If you imagine God as a Mother you would place yourself at 1. If you imagine God as a Father, you would place yourself at 7. If you imagine God as somewhere between Mother and Father, you would place yourself at 2, 3, 4, 5, or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F, SUBSTITUTING IMAGES A-F FOR "MOTHER" AND "FATHER".) Where would you place your image of God on the scale for . . . READ EACH SET OF IMAGES AND CIRCLE ONE CODE FOR EACH.
}
\section{Values}{

\itemize{
\item \code{1} creator
\item \code{7} healer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984, 1985, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/338/vshow}{GSS Data Explorer page for \code{crtrheal}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab creator \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab healer \tab no answer \tab don't know \tab Total \cr
   1984 \tab 484 \tab 310 \tab 42 \tab 35 \tab 427 \tab 27 \tab 26 \tab 70 \tab 52 \tab - \tab 1473 \cr
   1985 \tab - \tab 434 \tab 83 \tab 59 \tab 669 \tab 44 \tab 52 \tab 129 \tab 48 \tab 16 \tab 1534 \cr
   1986 \tab - \tab 432 \tab 66 \tab 48 \tab 665 \tab 35 \tab 40 \tab 131 \tab 45 \tab 8 \tab 1470 \cr
   Total \tab 484 \tab 1176 \tab 191 \tab 142 \tab 1761 \tab 106 \tab 118 \tab 330 \tab 145 \tab 24 \tab 4477 \cr
}
}

\seealso{
Other Core: 
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}},
\code{\link{rdeemlib}}
}
\concept{Core}
\keyword{variable}
