% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{cshutyp08}
\alias{cshutyp08}
\title{During april-june 2008, was r resident in household or institution}
\source{
General Social Survey https://gss.norc.org
}
\description{
cshutyp08
}
\details{
Question During April through June 2008, were you a resident in a household " such as living in a house, apartment or similar residence - or did you live elsewhere such as living in an institution like a college dorm, military barrack, work camp, shelter, residential treatment center, long-term medical care facility or nursing home?
}
\section{Values}{

\itemize{
\item \verb{[1]} living in household
\item \verb{[2]} living in institution
\item \verb{[3]} homeless (volunteered)
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5590/vshow}{GSS Data Explorer page for \code{cshutyp08}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab living in household \tab living in institution \tab homeless (volunteered) \tab other \tab no answer \tab don't know \tab Total \cr
   2010 \tab - \tab 2011 \tab 22 \tab 1 \tab 8 \tab 2 \tab - \tab 2044 \cr
   2012 \tab - \tab 1937 \tab 25 \tab 3 \tab 2 \tab 5 \tab 2 \tab 1974 \cr
   Total \tab 0 \tab 3948 \tab 47 \tab 4 \tab 10 \tab 7 \tab 2 \tab 4018 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp10}},
\code{\link{cshutyp12}}
}
\concept{Sampling and Weights}
\keyword{variable}
