% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{cshutyp12}
\alias{cshutyp12}
\title{During april-june 2012, was r resident in household or institution}
\source{
General Social Survey https://gss.norc.org
}
\description{
cshutyp12
}
\details{
Question During April through June 2012, were you a resident in a household – such as living in a house, apartment or similar residence - or did you live elsewhere such as living in an institution like a college dorm, military barrack, work camp, shelter, residential treatment center, long-term medical care facility or nursing home?
}
\section{Values}{

\itemize{
\item \code{1} living in household
\item \code{2} living in institution
\item \code{3} homeless (volunteered)
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5592/vshow}{GSS Data Explorer page for \code{cshutyp12}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{batch}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{subsamprate}},
\code{\link{version}},
\code{\link{vpsu}},
\code{\link{vstrat}},
\code{\link{wtss}},
\code{\link{wtssall}},
\code{\link{wtssnr}}
}
\concept{Sampling and Weights}
\keyword{variable}
