% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{curempyr}
\alias{curempyr}
\title{How many yrs r worked for current employer}
\source{
General Social Survey https://gss.norc.org
}
\description{
curempyr
}
\details{
Question How many years have you worked for your current employer?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3873/vshow}{GSS Data Explorer page for \code{curempyr}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 4 \tab 40 \tab 45 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 175 \tab 39 \tab 21 \tab 18 \tab 20 \tab 12 \tab 21 \tab 12 \tab 9 \tab 15 \tab 3 \tab 73 \tab 20 \tab 5 \tab 6 \tab 3 \tab 5 \tab 15 \tab 4 \tab 6 \tab 3 \tab 4 \tab 77 \tab 6 \tab 5 \tab 2 \tab 7 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 41 \tab 1 \tab 1 \tab 65 \tab 31 \tab 38 \tab 39 \tab 23 \tab 4 \tab 14 \tab 2023 \cr
   Total \tab 1165 \tab 175 \tab 39 \tab 21 \tab 18 \tab 20 \tab 12 \tab 21 \tab 12 \tab 9 \tab 15 \tab 3 \tab 73 \tab 20 \tab 5 \tab 6 \tab 3 \tab 5 \tab 15 \tab 4 \tab 6 \tab 3 \tab 4 \tab 77 \tab 6 \tab 5 \tab 2 \tab 7 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 41 \tab 1 \tab 1 \tab 65 \tab 31 \tab 38 \tab 39 \tab 23 \tab 4 \tab 14 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{whyleave}},
\code{\link{work3yrs}}
}
\concept{Global Economics}
\keyword{variable}
