% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_448.R
\name{cutdebt}
\alias{cutdebt}
\title{Should the government cut debt?}
\source{
General Social Survey https://gss.norc.org
}
\description{
cutdebt
}
\details{
Question Suppose the level of taxation in America stays the same as it is now. Should the government...
Please CHECK ONE BOX ONLY.
}
\section{Values}{

\itemize{
\item \verb{[1]} spend on social programs like health
\item \verb{[2]} spend less than now on social programs
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4107/vshow}{GSS Data Explorer page for \code{cutdebt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab spend on social programs like health \tab spend less than now on social programs \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 685 \tab 400 \tab 209 \tab 38 \tab 2904 \cr
   Total \tab 1572 \tab 685 \tab 400 \tab 209 \tab 38 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cctv}},
\code{\link{emonitor}},
\code{\link{govtinfo}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
