% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{cutgovt}
\alias{cutgovt}
\title{Cuts in govt spending}
\source{
General Social Survey https://gss.norc.org
}
\description{
cutgovt
}
\details{
Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it.
C. Cuts in government spending.
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor of
\item \code{2} in favor
\item \code{3} neither in favor nor against
\item \code{4} against
\item \code{5} strongly against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4051/vshow}{GSS Data Explorer page for \code{cutgovt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly in favor of \tab in favor \tab neither in favor nor against \tab against \tab strongly against \tab no answer \tab don't know \tab Total \cr
   1985 \tab 857 \tab 287 \tab 259 \tab 68 \tab 41 \tab 11 \tab 11 \tab - \tab 1534 \cr
   1990 \tab 155 \tab 436 \tab 488 \tab 186 \tab 52 \tab 20 \tab 34 \tab 1 \tab 1372 \cr
   1996 \tab 1572 \tab 535 \tab 543 \tab 139 \tab 53 \tab 23 \tab 37 \tab 2 \tab 2904 \cr
   2006 \tab 2992 \tab 410 \tab 529 \tab 333 \tab 147 \tab 64 \tab 2 \tab 33 \tab 4510 \cr
   2016 \tab 1477 \tab 413 \tab 471 \tab 294 \tab 143 \tab 54 \tab 8 \tab 7 \tab 2867 \cr
   Total \tab 7053 \tab 2081 \tab 2290 \tab 1020 \tab 436 \tab 172 \tab 92 \tab 43 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{poleff10}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{spenviro}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
