% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{cuthours}
\alias{cuthours}
\title{Reduction in length of work week}
\source{
General Social Survey https://gss.norc.org
}
\description{
cuthours
}
\details{
Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it.
H. Reducing the work week to create more jobs.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly in favor of
\item \verb{[2]} in favor
\item \verb{[3]} neither in favor nor against
\item \verb{[4]} against
\item \verb{[5]} strongly against
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4056/vshow}{GSS Data Explorer page for \code{cuthours}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly in favor of \tab in favor \tab neither in favor nor against \tab against \tab strongly against \tab no answer \tab don't know \tab Total \cr
   1985 \tab 857 \tab 44 \tab 139 \tab 174 \tab 229 \tab 84 \tab 7 \tab - \tab 1534 \cr
   1990 \tab 155 \tab 92 \tab 189 \tab 415 \tab 354 \tab 128 \tab 38 \tab 1 \tab 1372 \cr
   1996 \tab 1572 \tab 92 \tab 263 \tab 376 \tab 422 \tab 142 \tab 35 \tab 2 \tab 2904 \cr
   2006 \tab 2992 \tab 154 \tab 384 \tab 285 \tab 492 \tab 185 \tab 1 \tab 17 \tab 4510 \cr
   2016 \tab 1477 \tab 104 \tab 324 \tab 417 \tab 403 \tab 117 \tab 8 \tab 17 \tab 2867 \cr
   Total \tab 7053 \tab 486 \tab 1299 \tab 1667 \tab 1900 \tab 656 \tab 89 \tab 37 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutgovt}},
\code{\link{hlphitec}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{poleff10}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{spenviro}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
