% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_85.R
\name{cutspdr}
\alias{cutspdr}
\title{Cut govt spending--r's position}
\source{
General Social Survey https://gss.norc.org
}
\description{
cutspdr
}
\details{
Question Some people think the government should provide fewer services, even in areas such as health and education, in order to reduce spending. Other people feel it is important for the government to continue the services it now provides even if it means no reduction in spending. ENTER CODE IN BOXES.
A. Where would you place yourself on this scale, or haven't you thought much about this?
}
\section{Values}{

\itemize{
\item \code{0} havent thought much
\item \code{1} cut spending
\item \code{7} maintain services
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1980 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/763/vshow}{GSS Data Explorer page for \code{cutspdr}}.
}

\seealso{
Other Core: 
\code{\link{cutspdfg}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imprelax}},
\code{\link{impwork}}
}
\concept{Core}
\keyword{variable}
