% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_656.R
\name{cvdarpo}
\alias{cvdarpo}
\title{Change in archived livestreamed reading events since first year of covid}
\source{
General Social Survey https://gss.norc.org
}
\description{
cvdarpo
}
\details{
Question (Compared to the first year of the COVID-19 pandemic (March 2020 to March 2021), are you now...) watching or listening to archived web events featuring poetry, book-readings, or storytelling more often, at the same rate, less often, or are you not watching or listening to archived poetry, book-readings, or storytelling at all? This may have included video or audio clips on, for example, YouTube, an organization's website, or on a social media platform.

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \verb{[1]} more often
\item \verb{[2]} at the same rate
\item \verb{[3]} less often
\item \verb{[4]} r not attending these activities
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7888/vshow}{GSS Data Explorer page for \code{cvdarpo}}.

Counts by year:\tabular{lllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab Total \cr
   2022 \tab 2708 \tab 66 \tab 224 \tab 52 \tab 494 \tab 3544 \cr
   Total \tab 2708 \tab 66 \tab 224 \tab 52 \tab 494 \tab 3544 \cr
}
}

\seealso{
Other NEA: 
\code{\link{cvdclass}},
\code{\link{cvdpod}}
}
\concept{NEA}
\keyword{variable}
