% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_655.R
\name{cvdart}
\alias{cvdart}
\title{Change in in-person art since first year of covid}
\source{
General Social Survey https://gss.norc.org
}
\description{
cvdart
}
\details{
Question (Compared to the first year of the COVID-19 pandemic (March 2020 to March 2021), are you now...) attending in-person art exhibits, such as paintings, sculpture, textiles, graphic design, or photography, more often, at the same rate, less often, or are you not attending in-person art exhibits at all?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \verb{[1]} more often
\item \verb{[2]} at the same rate
\item \verb{[3]} less often
\item \verb{[4]} r not attending these activities
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7880/vshow}{GSS Data Explorer page for \code{cvdart}}.

Counts by year:\tabular{lllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab Total \cr
   2022 \tab 2703 \tab 149 \tab 194 \tab 101 \tab 397 \tab 3544 \cr
   Total \tab 2703 \tab 149 \tab 194 \tab 101 \tab 397 \tab 3544 \cr
}
}

\seealso{
Other NEA: 
\code{\link{cvdarmus}},
\code{\link{cvdcreat}},
\code{\link{cvdlvmus}},
\code{\link{cvdmov}},
\code{\link{cvdrdg}},
\code{\link{cvdstmus}},
\code{\link{cvdstpo}},
\code{\link{cvdtour}},
\code{\link{yrpod}}
}
\concept{NEA}
\keyword{variable}
