% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_655.R
\name{cvdmov}
\alias{cvdmov}
\title{Change in in-person movies since first year of covid}
\source{
General Social Survey https://gss.norc.org
}
\description{
cvdmov
}
\details{
Question (Compared to the first year of the COVID-19 pandemic (March 2020 to March 2021), are you now...) going to the movies more often, at the same rate, less often, or not going to the movies at all?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \verb{[1]} more often
\item \verb{[2]} at the same rate
\item \verb{[3]} less often
\item \verb{[4]} r not attending these activities
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7881/vshow}{GSS Data Explorer page for \code{cvdmov}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab more often \tab at the same rate \tab less often \tab r not attending these activities \tab don't know \tab iap \tab skipped on web \tab Total \cr
   2022 \tab - \tab 176 \tab 182 \tab 170 \tab 312 \tab 2 \tab 2701 \tab 1 \tab 3544 \cr
   Total \tab 0 \tab 176 \tab 182 \tab 170 \tab 312 \tab 2 \tab 2701 \tab 1 \tab 3544 \cr
}
}

\seealso{
Other NEA: 
\code{\link{cvdart}},
\code{\link{cvdcreat}},
\code{\link{cvdlvmus}},
\code{\link{cvdrdg}},
\code{\link{yrarmus}},
\code{\link{yrarpo}},
\code{\link{yrclass}},
\code{\link{yrpod}},
\code{\link{yrstpo}}
}
\concept{NEA}
\keyword{variable}
