% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{dailywrk}
\alias{dailywrk}
\title{How r's daily work is organized}
\source{
General Social Survey https://gss.norc.org
}
\description{
dailywrk
}
\details{
Question Which of the following statements best describes how your daily work is organized?
}
\section{Values}{

\itemize{
\item \code{1} i am free to decide how my daily work is organized
\item \code{2} i can decide within certain limits
\item \code{3} i am not free to decide how my daily work is organized
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4612/vshow}{GSS Data Explorer page for \code{dailywrk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab i am free to decide how my daily work is organized \tab i can decide within certain limits \tab i am not free to decide how my daily work is organized \tab no answer \tab can't choose \tab Total \cr
   2006 \tab 3492 \tab 325 \tab 485 \tab 206 \tab 2 \tab - \tab 4510 \cr
   2016 \tab 1927 \tab 375 \tab 423 \tab 130 \tab 8 \tab 4 \tab 2867 \cr
   Total \tab 5419 \tab 700 \tab 908 \tab 336 \tab 10 \tab 4 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
