% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{danger}
\alias{danger}
\title{How often does r work in dangerous conds}
\source{
General Social Survey https://gss.norc.org
}
\description{
danger
}
\details{
Question Now some more questions about your working conditions. one code for each item below to show how often it applies to your work. How often...
E. Do you work in dangerous conditions?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} hardly ever
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4554/vshow}{GSS Data Explorer page for \code{danger}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 36 \tab 71 \tab 176 \tab 192 \tab 396 \tab 3 \tab 24 \tab 1537 \cr
   1998 \tab 1967 \tab 51 \tab 55 \tab 143 \tab 226 \tab 350 \tab 3 \tab 37 \tab 2832 \cr
   2006 \tab 3492 \tab 74 \tab 85 \tab 154 \tab 213 \tab 490 \tab - \tab 2 \tab 4510 \cr
   Total \tab 6098 \tab 161 \tab 211 \tab 473 \tab 631 \tab 1236 \tab 6 \tab 63 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
