% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_224.R
\name{dangoth1}
\alias{dangoth1}
\title{How dangerous r thinks person 1 is to others}
\source{
General Social Survey https://gss.norc.org
}
\description{
dangoth1
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very, in your opinion, how dangerous is (NAME 1) to others?
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all dangerous
\item \code{10} 10 - very dangerous
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7135/vshow}{GSS Data Explorer page for \code{dangoth1}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all dangerous \tab 10 - very dangerous \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1751 \tab 307 \tab 23 \tab 72 \tab 55 \tab 29 \tab 30 \tab 26 \tab 19 \tab 15 \tab 7 \tab 11 \tab 3 \tab 2348 \cr
   Total \tab 1751 \tab 307 \tab 23 \tab 72 \tab 55 \tab 29 \tab 30 \tab 26 \tab 19 \tab 15 \tab 7 \tab 11 \tab 3 \tab 2348 \cr
}
}

\keyword{variable}
