% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{dangrslf}
\alias{dangrslf}
\title{X should be admitted to be hospitalized for treatment if dangerous to x}
\source{
General Social Survey https://gss.norc.org
}
\description{
dangrslf
}
\details{
Question Some cities and states have laws that force people with problems like NAME into treatment. Do you think that people like NAME should be forced by law to...
D. Be admitted to a hospital for treatment if s/he is dangerous to him/herself.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2034/vshow}{GSS Data Explorer page for \code{dangrslf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 1227 \tab 158 \tab 38 \tab 21 \tab 2904 \cr
   2006 \tab 3073 \tab 1224 \tab 165 \tab 31 \tab 17 \tab 4510 \cr
   2018 \tab 1175 \tab 1075 \tab 54 \tab 15 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 3526 \tab 377 \tab 84 \tab 67 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mntlas1}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}},
\code{\link{spmentl}}
}
\concept{Mental Health}
\keyword{variable}
