% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_225.R
\name{dangslf1}
\alias{dangslf1}
\title{How dangerous r thinks person 1 is to themself}
\source{
General Social Survey https://gss.norc.org
}
\description{
dangslf1
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very, in your opinion, how dangerous is (NAME 1) to himself/herself?
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all dangerous
\item \code{10} 10 - very dangerous
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7140/vshow}{GSS Data Explorer page for \code{dangslf1}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all dangerous \tab 10 - very dangerous \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1751 \tab 185 \tab 55 \tab 68 \tab 43 \tab 39 \tab 48 \tab 42 \tab 47 \tab 41 \tab 15 \tab 12 \tab 2 \tab 2348 \cr
   Total \tab 1751 \tab 185 \tab 55 \tab 68 \tab 43 \tab 39 \tab 48 \tab 42 \tab 47 \tab 41 \tab 15 \tab 12 \tab 2 \tab 2348 \cr
}
}

\keyword{variable}
