% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_225.R
\name{dangslf3}
\alias{dangslf3}
\title{How dangerous r thinks person 3 is to themself}
\source{
General Social Survey https://gss.norc.org
}
\description{
dangslf3
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very, in your opinion, how dangerous is (NAME 3) to himself/herself?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - not at all dangerous
\item \verb{[10]} 10 - very dangerous
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7142/vshow}{GSS Data Explorer page for \code{dangslf3}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all dangerous \tab 10 - very dangerous \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2095 \tab 104 \tab 14 \tab 32 \tab 22 \tab 16 \tab 16 \tab 7 \tab 16 \tab 12 \tab 5 \tab 2 \tab 7 \tab 2348 \cr
   Total \tab 2095 \tab 104 \tab 14 \tab 32 \tab 22 \tab 16 \tab 16 \tab 7 \tab 16 \tab 12 \tab 5 \tab 2 \tab 7 \tab 2348 \cr
}
}

\keyword{variable}
