% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_225.R
\name{dangslf5}
\alias{dangslf5}
\title{How dangerous r thinks person 5 is to themself}
\source{
General Social Survey https://gss.norc.org
}
\description{
dangslf5
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very, in your opinion, how dangerous is (NAME 5) to himself/herself?
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all dangerous
\item \code{10} 10 - very dangerous
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7144/vshow}{GSS Data Explorer page for \code{dangslf5}}.
}

\keyword{variable}
