% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{databank}
\alias{databank}
\title{Computer data threat to individual privacy}
\source{
General Social Survey https://gss.norc.org
}
\description{
databank
}
\details{
Question The federal government has a lot of different pieces of information about people which computers can bring together very quickly. Is this a very serious threat to individual privacy, a fairly serious threat, not a serious threat, or not a threat at all to individual privacy?
}
\section{Values}{

\itemize{
\item \code{1} very serious threat
\item \code{2} fairly serious
\item \code{3} not serious
\item \code{4} not a threat at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4003/vshow}{GSS Data Explorer page for \code{databank}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very serious threat \tab fairly serious \tab not serious \tab not a threat at all \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 197 \tab 191 \tab 176 \tab 56 \tab 46 \tab 11 \tab 1534 \cr
   1996 \tab 1572 \tab 478 \tab 422 \tab 254 \tab 57 \tab 93 \tab 28 \tab 2904 \cr
   2006 \tab 2992 \tab 398 \tab 593 \tab 397 \tab 104 \tab 23 \tab 3 \tab 4510 \cr
   Total \tab 5421 \tab 1073 \tab 1206 \tab 827 \tab 217 \tab 162 \tab 42 \tab 8948 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{progtax}},
\code{\link{verdict}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
