% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{dataprot}
\alias{dataprot}
\title{Responsibility for internet data protection}
\source{
General Social Survey https://gss.norc.org
}
\description{
dataprot
}
\details{
Question Who should be mostly responsible for protecting personal information when using the internet?
}
\section{Values}{

\itemize{
\item \verb{[1]} online companies - as they need to ensure they process personal information safely
\item \verb{[2]} people - as they need to take care of their personal information
\item \verb{[3]} public authorities - as they need to ensure that personal information is protected
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8103/vshow}{GSS Data Explorer page for \code{dataprot}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab online companies - as they need to ensure they process personal information safely \tab people - as they need to take care of their personal information \tab public authorities - as they need to ensure that personal information is protected \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 542 \tab 593 \tab 212 \tab 271 \tab 1673 \tab 4 \tab 14 \tab 3309 \cr
   Total \tab 0 \tab 542 \tab 593 \tab 212 \tab 271 \tab 1673 \tab 4 \tab 14 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{aidrive}},
\code{\link{aimed}},
\code{\link{aiworry}},
\code{\link{classtech}},
\code{\link{eductech}},
\code{\link{emonitor1}},
\code{\link{ftruse}},
\code{\link{gendtech}},
\code{\link{harmgood1}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{intaccprxy}},
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{intcnfrm}},
\code{\link{intcomm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{inthome}},
\code{\link{inthrss}},
\code{\link{intliteracy}},
\code{\link{intlnly}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intrnetuse}},
\code{\link{intrust}},
\code{\link{intscam}},
\code{\link{intsearch}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{intstream}},
\code{\link{intsurf}},
\code{\link{inttrig}},
\code{\link{intviews}},
\code{\link{modelobby}},
\code{\link{modeorg}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{nextgen1}},
\code{\link{nouseint1}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint2}},
\code{\link{nouseint3}},
\code{\link{nouseint4}},
\code{\link{nouseint5}},
\code{\link{nouseint6}},
\code{\link{nouseint7}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{papernews}},
\code{\link{polnewsfrom}},
\code{\link{prxyuse}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{techesy}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{tvnews1}},
\code{\link{unplgrsn}},
\code{\link{unplug}},
\code{\link{urbantech}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
