% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_690.R
\name{dateintv_next}
\alias{dateintv_next}
\title{Date of gss next interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
dateintv_next
}
\details{
Question Date of GSS Next interview
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8258/vshow}{GSS Data Explorer page for \code{dateintv_next}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 1001 \tab 1002 \tab 1003 \tab 1004 \tab 1005 \tab 1006 \tab 1007 \tab 1008 \tab 1009 \tab 1010 \tab 1011 \tab 1013 \tab 1014 \tab 1015 \tab 1016 \tab 1017 \tab 1018 \tab 1019 \tab 1020 \tab 1021 \tab 820 \tab 821 \tab 822 \tab 823 \tab 824 \tab 825 \tab 826 \tab 827 \tab 828 \tab 829 \tab 830 \tab 831 \tab 901 \tab 902 \tab 903 \tab 904 \tab 905 \tab 906 \tab 907 \tab 908 \tab 909 \tab 910 \tab 911 \tab 912 \tab 913 \tab 914 \tab 915 \tab 916 \tab 917 \tab 918 \tab 919 \tab 920 \tab 921 \tab 923 \tab 924 \tab 925 \tab 926 \tab 927 \tab 928 \tab 929 \tab 930 \tab iap \tab Total \cr
   2024 \tab - \tab 25 \tab 13 \tab 25 \tab 11 \tab 4 \tab 9 \tab 23 \tab 15 \tab 5 \tab 2 \tab 1 \tab 5 \tab 13 \tab 29 \tab 17 \tab 15 \tab 9 \tab 9 \tab 9 \tab 6 \tab 20 \tab 36 \tab 35 \tab 46 \tab 44 \tab 75 \tab 46 \tab 30 \tab 14 \tab 70 \tab 26 \tab 16 \tab 12 \tab 17 \tab 26 \tab 3 \tab 8 \tab 5 \tab 4 \tab 8 \tab 24 \tab 32 \tab 7 \tab 3 \tab 6 \tab 2 \tab 8 \tab 3 \tab 31 \tab 11 \tab 8 \tab 1 \tab 3 \tab 7 \tab 3 \tab 26 \tab 8 \tab 21 \tab 22 \tab 27 \tab 26 \tab 2244 \tab 3309 \cr
   Total \tab 0 \tab 25 \tab 13 \tab 25 \tab 11 \tab 4 \tab 9 \tab 23 \tab 15 \tab 5 \tab 2 \tab 1 \tab 5 \tab 13 \tab 29 \tab 17 \tab 15 \tab 9 \tab 9 \tab 9 \tab 6 \tab 20 \tab 36 \tab 35 \tab 46 \tab 44 \tab 75 \tab 46 \tab 30 \tab 14 \tab 70 \tab 26 \tab 16 \tab 12 \tab 17 \tab 26 \tab 3 \tab 8 \tab 5 \tab 4 \tab 8 \tab 24 \tab 32 \tab 7 \tab 3 \tab 6 \tab 2 \tab 8 \tab 3 \tab 31 \tab 11 \tab 8 \tab 1 \tab 3 \tab 7 \tab 3 \tab 26 \tab 8 \tab 21 \tab 22 \tab 27 \tab 26 \tab 2244 \tab 3309 \cr
}
}

\concept{Date Of Interview}
\concept{Date of Interview}
\keyword{variable}
