% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{datesch}
\alias{datesch}
\title{Date schooling completed}
\source{
General Social Survey https://gss.norc.org
}
\description{
datesch
}
\details{
Question In what month and year did you complete that year of school? RESPONDENT'S DATE OF FINISHING EDUCATION
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1460/vshow}{GSS Data Explorer page for \code{datesch}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1205 \tab 1206 \tab 1604 \tab 1605 \tab 2112 \tab 2206 \tab 2305 \tab 2405 \tab 2499 \tab 2505 \tab 2506 \tab 2605 \tab 2609 \tab 2699 \tab 2705 \tab 2706 \tab 2805 \tab 2806 \tab 2898 \tab 2904 \tab 2906 \tab 3005 \tab 3006 \tab 3099 \tab 3105 \tab 3205 \tab 3206 \tab 3301 \tab 3305 \tab 3306 \tab 3405 \tab 3406 \tab 3504 \tab 3505 \tab 3506 \tab 3605 \tab 3606 \tab 3698 \tab 3699 \tab 3705 \tab 3706 \tab 3798 \tab 3805 \tab 3806 \tab 3905 \tab 3906 \tab 4005 \tab 4006 \tab 4104 \tab 4105 \tab 4106 \tab 4202 \tab 4205 \tab 4206 \tab 4301 \tab 4304 \tab 4305 \tab 4306 \tab 4309 \tab 4398 \tab 4401 \tab 4405 \tab 4406 \tab 4499 \tab 4503 \tab 4505 \tab 4506 \tab 4507 \tab 4508 \tab 4510 \tab 4511 \tab 4598 \tab 4602 \tab 4603 \tab 4605 \tab 4606 \tab 4612 \tab 4705 \tab 4706 \tab 4801 \tab 4805 \tab 4806 \tab 4809 \tab 4812 \tab 4898 \tab 4901 \tab 4903 \tab 4904 \tab 4905 \tab 4906 \tab 4998 \tab 5004 \tab 5005 \tab 5006 \tab 5007 \tab 5098 \tab 5105 \tab 5106 \tab 5108 \tab 5201 \tab 5205 \tab 5206 \tab 5305 \tab 5306 \tab 5398 \tab 5404 \tab 5405 \tab 5406 \tab 5409 \tab 5411 \tab 5501 \tab 5505 \tab 5506 \tab 5508 \tab 5598 \tab 5601 \tab 5605 \tab 5606 \tab 5610 \tab 5698 \tab 5705 \tab 5706 \tab 5707 \tab 5710 \tab 5712 \tab 5798 \tab 5805 \tab 5806 \tab 5905 \tab 5906 \tab 5912 \tab 6005 \tab 6006 \tab 6007 \tab 6098 \tab 6099 \tab 6103 \tab 6104 \tab 6105 \tab 6106 \tab 6108 \tab 6205 \tab 6206 \tab 6208 \tab 6298 \tab 6302 \tab 6305 \tab 6306 \tab 6307 \tab 6310 \tab 6405 \tab 6406 \tab 6407 \tab 6411 \tab 6498 \tab 6502 \tab 6503 \tab 6505 \tab 6506 \tab 6508 \tab 6599 \tab 6605 \tab 6606 \tab 6608 \tab 6611 \tab 6702 \tab 6705 \tab 6706 \tab 6709 \tab 6712 \tab 6801 \tab 6805 \tab 6806 \tab 6809 \tab 6898 \tab 6901 \tab 6905 \tab 6906 \tab 6910 \tab 6911 \tab 6912 \tab 7001 \tab 7005 \tab 7006 \tab 7008 \tab 7010 \tab 7011 \tab 7012 \tab 7103 \tab 7105 \tab 7106 \tab 7107 \tab 7109 \tab 7111 \tab 7112 \tab 7198 \tab 7205 \tab 7206 \tab 7208 \tab 7301 \tab 7302 \tab 7304 \tab 7305 \tab 7306 \tab 7308 \tab 7398 \tab 7401 \tab 7403 \tab 7405 \tab 7406 \tab 7407 \tab 7408 \tab 7409 \tab 7412 \tab 7498 \tab 7501 \tab 7505 \tab 7506 \tab 7507 \tab 7508 \tab 7509 \tab 7512 \tab 7598 \tab 7604 \tab 7605 \tab 7606 \tab 7607 \tab 7608 \tab 7609 \tab 7612 \tab 7698 \tab 7699 \tab 7701 \tab 7703 \tab 7705 \tab 7706 \tab 7707 \tab 7712 \tab 7798 \tab 7801 \tab 7805 \tab 7806 \tab 7807 \tab 7808 \tab 7812 \tab 7898 \tab 7902 \tab 7904 \tab 7905 \tab 7906 \tab 7908 \tab 7909 \tab 7910 \tab 7912 \tab 7998 \tab 8001 \tab 8002 \tab 8005 \tab 8006 \tab 8007 \tab 8009 \tab 8010 \tab 8012 \tab 8098 \tab 8102 \tab 8105 \tab 8106 \tab 8108 \tab 8112 \tab 8198 \tab 8201 \tab 8203 \tab 8204 \tab 8205 \tab 8206 \tab 8207 \tab 8212 \tab 8304 \tab 8305 \tab 8306 \tab 8308 \tab 8310 \tab 8401 \tab 8403 \tab 8404 \tab 8405 \tab 8406 \tab 8407 \tab 8409 \tab 8411 \tab 8412 \tab 8503 \tab 8504 \tab 8505 \tab 8506 \tab 8508 \tab 8509 \tab 8511 \tab 8512 \tab 8598 \tab 8601 \tab 8603 \tab 8604 \tab 8605 \tab 8606 \tab 8608 \tab 8609 \tab 8610 \tab 8612 \tab 8698 \tab 8701 \tab 8702 \tab 8703 \tab 8704 \tab 8705 \tab 8706 \tab 8707 \tab 8708 \tab 8712 \tab 8803 \tab 8805 \tab 8806 \tab 8811 \tab 8812 \tab 8903 \tab 8904 \tab 8905 \tab 8906 \tab 8908 \tab 8909 \tab 8912 \tab 8998 \tab 9001 \tab 9004 \tab 9005 \tab 9006 \tab 9007 \tab 9008 \tab 9009 \tab 9012 \tab 9101 \tab 9103 \tab 9104 \tab 9105 \tab 9106 \tab 9107 \tab 9108 \tab 9110 \tab 9112 \tab 9198 \tab 9202 \tab 9203 \tab 9204 \tab 9205 \tab 9206 \tab 9207 \tab 9208 \tab 9209 \tab 9211 \tab 9212 \tab 9301 \tab 9303 \tab 9304 \tab 9305 \tab 9306 \tab 9307 \tab 9308 \tab 9309 \tab 9311 \tab 9312 \tab 9398 \tab 9399 \tab 9401 \tab 9403 \tab 9404 \tab 9406 \tab 9805 \tab 9806 \tab 9898 \tab 9906 \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1474 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 5 \tab 3 \tab 6 \tab 1 \tab 3 \tab 4 \tab 4 \tab 7 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 9 \tab 4 \tab 6 \tab 3 \tab 5 \tab 2 \tab 5 \tab 6 \tab 1 \tab 3 \tab 5 \tab 2 \tab 1 \tab 4 \tab 6 \tab 2 \tab 2 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab 4 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 3 \tab 7 \tab 1 \tab 2 \tab 9 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 1 \tab 5 \tab 7 \tab 1 \tab 1 \tab 4 \tab 7 \tab 1 \tab 1 \tab 3 \tab 9 \tab 8 \tab 6 \tab 1 \tab 1 \tab 4 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 12 \tab 2 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 6 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 6 \tab 7 \tab 6 \tab 2 \tab 7 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 1 \tab 3 \tab 14 \tab 4 \tab 3 \tab 1 \tab 2 \tab 9 \tab 1 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 7 \tab 10 \tab 1 \tab 1 \tab 5 \tab 8 \tab 1 \tab 1 \tab 2 \tab 3 \tab 9 \tab 2 \tab 1 \tab 1 \tab 3 \tab 13 \tab 2 \tab 1 \tab 2 \tab 8 \tab 14 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 13 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 18 \tab 1 \tab 1 \tab 3 \tab 1 \tab 7 \tab 15 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 13 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 10 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 14 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 9 \tab 14 \tab 2 \tab 4 \tab 2 \tab 1 \tab 14 \tab 17 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 15 \tab 17 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 15 \tab 21 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 9 \tab 18 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 11 \tab 20 \tab 3 \tab 2 \tab 1 \tab 7 \tab 17 \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 19 \tab 13 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 9 \tab 10 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 9 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 16 \tab 10 \tab 1 \tab 2 \tab 2 \tab 1 \tab 9 \tab 11 \tab 1 \tab 5 \tab 1 \tab 1 \tab 9 \tab 8 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab 1 \tab 17 \tab 6 \tab 1 \tab 3 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 14 \tab 11 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 12 \tab 21 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 35 \tab 28 \tab 1 \tab 4 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 2 \tab 36 \tab 1 \tab 20 \tab - \tab 2992 \cr
   Total \tab 1474 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 5 \tab 3 \tab 6 \tab 1 \tab 3 \tab 4 \tab 4 \tab 7 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 9 \tab 4 \tab 6 \tab 3 \tab 5 \tab 2 \tab 5 \tab 6 \tab 1 \tab 3 \tab 5 \tab 2 \tab 1 \tab 4 \tab 6 \tab 2 \tab 2 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab 4 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 3 \tab 7 \tab 1 \tab 2 \tab 9 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 1 \tab 5 \tab 7 \tab 1 \tab 1 \tab 4 \tab 7 \tab 1 \tab 1 \tab 3 \tab 9 \tab 8 \tab 6 \tab 1 \tab 1 \tab 4 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 12 \tab 2 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 6 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 6 \tab 7 \tab 6 \tab 2 \tab 7 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 1 \tab 3 \tab 14 \tab 4 \tab 3 \tab 1 \tab 2 \tab 9 \tab 1 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 7 \tab 10 \tab 1 \tab 1 \tab 5 \tab 8 \tab 1 \tab 1 \tab 2 \tab 3 \tab 9 \tab 2 \tab 1 \tab 1 \tab 3 \tab 13 \tab 2 \tab 1 \tab 2 \tab 8 \tab 14 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 13 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 18 \tab 1 \tab 1 \tab 3 \tab 1 \tab 7 \tab 15 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 13 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 10 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 14 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 9 \tab 14 \tab 2 \tab 4 \tab 2 \tab 1 \tab 14 \tab 17 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 15 \tab 17 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 15 \tab 21 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 9 \tab 18 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 11 \tab 20 \tab 3 \tab 2 \tab 1 \tab 7 \tab 17 \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 19 \tab 13 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 9 \tab 10 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 9 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 16 \tab 10 \tab 1 \tab 2 \tab 2 \tab 1 \tab 9 \tab 11 \tab 1 \tab 5 \tab 1 \tab 1 \tab 9 \tab 8 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab 1 \tab 17 \tab 6 \tab 1 \tab 3 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 14 \tab 11 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 12 \tab 21 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 35 \tab 28 \tab 1 \tab 4 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 2 \tab 36 \tab 1 \tab 20 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{datefrst}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{fstspyr}},
\code{\link{indfirst}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpicked}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{malive1}},
\code{\link{marnum}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{palive1}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind10}},
\code{\link{sbind80}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sbocc80}},
\code{\link{sboccindv}},
\code{\link{sboccstatus}},
\code{\link{sbocctag}},
\code{\link{sbpicked}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{spfam16}},
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre}},
\code{\link{sphedpre10}},
\code{\link{sphedpre105plus}},
\code{\link{sphedslf}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spsibs}},
\code{\link{spyrborn}},
\code{\link{wrkslfst}}

Other Education: 
\code{\link{actgreek}},
\code{\link{actschl}},
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{bilinged}},
\code{\link{cathsch}},
\code{\link{chldeduc}},
\code{\link{chrissch}},
\code{\link{classics}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{coldeg1}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{colop}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{degree}},
\code{\link{denomsch}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{educ}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{educemp}},
\code{\link{educop}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{govedop}},
\code{\link{homesch}},
\code{\link{hsbasics}},
\code{\link{hsbio}},
\code{\link{hscaring}},
\code{\link{hschem}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsmath}},
\code{\link{hsorder}},
\code{\link{hsphys}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inteduc}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{majorcol}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memschl}},
\code{\link{natdeg}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{opeduc}},
\code{\link{oppared}},
\code{\link{othsch}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{relsch}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{schleth}},
\code{\link{schlhead}},
\code{\link{schoolng}},
\code{\link{scimath}},
\code{\link{sexeduc}},
\code{\link{solgreek}},
\code{\link{solschl}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{speduc}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{teacheth}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{trstprof}},
\code{\link{undoccol}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{whoteach}}
}
\concept{Education}
\concept{Family Mobility}
\keyword{variable}
