% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{daunum}
\alias{daunum}
\title{Number of adult daughter's}
\source{
General Social Survey https://gss.norc.org
}
\description{
daunum
}
\details{
Question How many daughters age 18 or older do you have? (We mean daughters who are still alive; please include step-daughters, half-daughters and adopted daughters.)
}
\section{Values}{

\itemize{
\item \code{0} none
\item \code{1} one
\item \code{2} two
\item \code{3} three
\item \code{4} four
\item \code{5} five +
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4143/vshow}{GSS Data Explorer page for \code{daunum}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab none \tab one \tab two \tab three \tab four \tab five + \tab no answer \tab Total \cr
   1986 \tab - \tab 978 \tab 254 \tab 132 \tab 63 \tab 23 \tab 17 \tab 3 \tab 1470 \cr
   Total \tab 0 \tab 978 \tab 254 \tab 132 \tab 63 \tab 23 \tab 17 \tab 3 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{brocall}},
\code{\link{daucall}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{posslq}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
