% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{dauvisit}
\alias{dauvisit}
\title{How often visit with daughter}
\source{
General Social Survey https://gss.norc.org
}
\description{
dauvisit
}
\details{
Question How many daughters age 18 or older do you have? (We mean daughters who are still alive; please include step-daughters, half-daughters and adopted daughters.)
A. How often do you see or visit with your daughter?
}
\section{Values}{

\itemize{
\item \code{1} lives in hshld
\item \code{2} daily
\item \code{3} several x week
\item \code{4} once a week
\item \code{5} once a month
\item \code{6} several x year
\item \code{7} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4144/vshow}{GSS Data Explorer page for \code{dauvisit}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab lives in hshld \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab no answer \tab Total \cr
   1986 \tab 978 \tab 69 \tab 66 \tab 78 \tab 78 \tab 71 \tab 70 \tab 57 \tab 3 \tab 1470 \cr
   Total \tab 978 \tab 69 \tab 66 \tab 78 \tab 78 \tab 71 \tab 70 \tab 57 \tab 3 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{brocall}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{posslq}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
