% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_478.R
\name{daycare1}
\alias{daycare1}
\title{Public daycare center ok for kid ( 3 yrs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
daycare1
}
\details{
Question Think of a child under 3 years old whose parents both have full-time jobs. How suitable do you think each of these child care arrangements would be for the child?
A. A public day care center?
}
\section{Values}{

\itemize{
\item \code{1} very suitable
\item \code{2} somewhat suitable
\item \code{3} not very suitable
\item \code{4} not at all suitable
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4357/vshow}{GSS Data Explorer page for \code{daycare1}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{ilikejob}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
