% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_478.R
\name{daycare4}
\alias{daycare4}
\title{Friend or neighbor ok for kid ( 3 yrs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
daycare4
}
\details{
Question Think of a child under 3 years old whose parents both have full-time jobs. How suitable do you think each of these child care arrangements would be for the child?
D. A neighbor or friend?
}
\section{Values}{

\itemize{
\item \verb{[1]} very suitable
\item \verb{[2]} somewhat suitable
\item \verb{[3]} not very suitable
\item \verb{[4]} not at all suitable
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4360/vshow}{GSS Data Explorer page for \code{daycare4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very suitable \tab somewhat suitable \tab not very suitable \tab not at all suitable \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 367 \tab 607 \tab 220 \tab 84 \tab 81 \tab 55 \tab 1481 \cr
   Total \tab 67 \tab 367 \tab 607 \tab 220 \tab 84 \tab 81 \tab 55 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare5}},
\code{\link{ilikejob}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
