% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{daynight}
\alias{daynight}
\title{When did most errors occur}
\source{
General Social Survey https://gss.norc.org
}
\description{
daynight
}
\details{
Question Please look at Card I-3. Day-night rhythms dramatically affect our bodies. Probably no body system is more  influenced than the nervous system. The figure on Card I-3 illustrates the number of errors made by shift workers in different portions of the 24-hour cycle. Based on the data illustrated in the figure, during which of these time periods did the most errors occur?
}
\section{Values}{

\itemize{
\item \code{1} 2 a.m. to 4 a.m.
\item \code{2} 8 a.m. to 10 a.m.
\item \code{3} 12 p.m. to 2 p.m.
\item \code{4} 2 p.m. to 4 p.m.
\item \code{5} 8 p.m. to 10 p.m.
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3482/vshow}{GSS Data Explorer page for \code{daynight}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}

Other Knowledge: 
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}

Other Science: 
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
