% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{dclindus}
\alias{dclindus}
\title{Govt should support declining industries to protect jobs}
\source{
General Social Survey https://gss.norc.org
}
\description{
dclindus
}
\details{
Question The government should support declining industries to protect jobs, even if it might require a tax increase to pay for it
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7553/vshow}{GSS Data Explorer page for \code{dclindus}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2246 \tab 54 \tab 311 \tab 582 \tab 553 \tab 177 \tab 82 \tab 2 \tab 25 \tab 4032 \cr
   Total \tab 2246 \tab 54 \tab 311 \tab 582 \tab 553 \tab 177 \tab 82 \tab 2 \tab 25 \tab 4032 \cr
}
}

\keyword{variable}
