% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{death}
\alias{death}
\title{R's possible vital status as of 2014}
\source{
General Social Survey https://gss.norc.org
}
\description{
death
}
\details{
Question Respondent's possible vital status as of 2014
}
\section{Values}{

\itemize{
\item \code{1} possibly alive
\item \code{2} possibly dead
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7426/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab possibly alive \tab possibly dead \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 14 \tab 748 \tab 770 \tab - \tab 1532 \cr
   1980 \tab 178 \tab 645 \tab 645 \tab - \tab 1468 \cr
   1982 \tab 125 \tab 887 \tab 848 \tab - \tab 1860 \cr
   1983 \tab 239 \tab 742 \tab 618 \tab - \tab 1599 \cr
   1984 \tab 50 \tab 781 \tab 642 \tab - \tab 1473 \cr
   1985 \tab 57 \tab 741 \tab 736 \tab - \tab 1534 \cr
   1986 \tab 88 \tab 733 \tab 649 \tab - \tab 1470 \cr
   1987 \tab 81 \tab 1004 \tab 734 \tab - \tab 1819 \cr
   1988 \tab 17 \tab 867 \tab 597 \tab - \tab 1481 \cr
   1989 \tab 44 \tab 898 \tab 595 \tab - \tab 1537 \cr
   1990 \tab 15 \tab 843 \tab 514 \tab - \tab 1372 \cr
   1991 \tab 24 \tab 955 \tab 538 \tab - \tab 1517 \cr
   1993 \tab 57 \tab 1106 \tab 443 \tab - \tab 1606 \cr
   1994 \tab 35 \tab 2159 \tab 798 \tab - \tab 2992 \cr
   1996 \tab 60 \tab 2176 \tab 668 \tab - \tab 2904 \cr
   1998 \tab 116 \tab 2129 \tab 587 \tab - \tab 2832 \cr
   2000 \tab 154 \tab 2134 \tab 529 \tab - \tab 2817 \cr
   2002 \tab 188 \tab 2134 \tab 443 \tab - \tab 2765 \cr
   2004 \tab 82 \tab 2344 \tab 386 \tab - \tab 2812 \cr
   2006 \tab 95 \tab 3898 \tab 517 \tab - \tab 4510 \cr
   2008 \tab 46 \tab 1798 \tab 179 \tab - \tab 2023 \cr
   2010 \tab 28 \tab 1894 \tab 122 \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 24672 \tab 31616 \tab 12558 \tab 6853 \tab 75699 \cr
}
}

\keyword{variable}
