% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{death16}
\alias{death16}
\title{Relatives died since age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
death16
}
\details{
Question B. Number of relatives dying since respondent 16 years old and prior to the last five years.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/747/vshow}{GSS Data Explorer page for \code{death16}}.
}

\seealso{
Other Core: 
\code{\link{chlddth}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{madeath}},
\code{\link{padeath}},
\code{\link{unrel1}},
\code{\link{unrel4}}

Other Family: 
\code{\link{chlddth}},
\code{\link{death5}},
\code{\link{madeath}},
\code{\link{padeath}}

Other Relatives: 
\code{\link{death5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{unrel1}},
\code{\link{unrel4}}

Other Trauma: 
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{unrel1}},
\code{\link{unrel4}}
}
\concept{Core}
\concept{Family}
\concept{Relatives}
\concept{Trauma}
\keyword{variable}
