% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{deathpen}
\alias{deathpen}
\title{Murderers should get death penalty}
\source{
General Social Survey https://gss.norc.org
}
\description{
deathpen
}
\details{
Question Here are some measures to deal with crime. Some people are in favor of them while other people are against them. Do you agree or disagree that. . .
B. People convicted of murder should be subject to the death penalty.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4629/vshow}{GSS Data Explorer page for \code{deathpen}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 642 \tab 327 \tab 104 \tab 138 \tab 84 \tab 59 \tab 5 \tab 1517 \cr
   Total \tab 158 \tab 642 \tab 327 \tab 104 \tab 138 \tab 84 \tab 59 \tab 5 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{hapunhap}},
\code{\link{premars1}},
\code{\link{stiffpun}},
\code{\link{xmarsex1}}
}
\concept{ISSP Religion}
\keyword{variable}
