% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_200.R
\name{decgift}
\alias{decgift}
\title{Deciding on major charitable gifts.}
\source{
General Social Survey https://gss.norc.org
}
\description{
decgift
}
\details{
Question This question is about some of the financial decisions that you and your (husband/wife) may make. For each decision, please tell me how it was made the last time you faced this decision.
C. Deciding on major charitable gifts.
}
\section{Values}{

\itemize{
\item \verb{[1]} husband alone
\item \verb{[2]} husband after consulting with wife
\item \verb{[3]} husband and wife together
\item \verb{[4]} wife after consulting husband
\item \verb{[5]} wife alone
\item \verb{[6]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1883/vshow}{GSS Data Explorer page for \code{decgift}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab husband alone \tab husband after consulting with wife \tab husband and wife together \tab wife after consulting husband \tab wife alone \tab other \tab no answer \tab see codebook \tab Total \cr
   1996 \tab 2196 \tab 29 \tab 29 \tab 475 \tab 45 \tab 53 \tab 15 \tab 6 \tab 56 \tab 2904 \cr
   Total \tab 2196 \tab 29 \tab 29 \tab 475 \tab 45 \tab 53 \tab 15 \tab 6 \tab 56 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decbed}},
\code{\link{layoffs}}
}
\concept{Markets}
\keyword{variable}
