% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{decidwrk}
\alias{decidwrk}
\title{How are your working hours decided}
\source{
General Social Survey https://gss.norc.org
}
\description{
decidwrk
}
\details{
Question (HAND CARD D9)
Which of the following statements best describes how your working hours are decided?
(By working hours we mean here the times you start and finish work, and not the total hours you work per week or month.)
}
\section{Values}{

\itemize{
\item \verb{[1]} starting and finishing times are decided by my employer and i cannot change them on my own
\item \verb{[2]} i have a schedule or shift which regularly changes (for example, from days to evenings or to nights)
\item \verb{[3]} i have a schedule where daily working times are decided at short notice by my employer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6332/vshow}{GSS Data Explorer page for \code{decidwrk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab starting and finishing times are decided by my employer and i cannot change them on my own \tab i have a schedule or shift which regularly changes (for example, from days to evenings or to nights) \tab i have a schedule where daily working times are decided at short notice by my employer \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 445 \tab 383 \tab 104 \tab 2 \tab 6 \tab 2867 \cr
   Total \tab 1927 \tab 445 \tab 383 \tab 104 \tab 2 \tab 6 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
