% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{decself}
\alias{decself}
\title{Imp of personal judgment to make decision}
\source{
General Social Survey https://gss.norc.org
}
\description{
decself
}
\details{
Question Can you tell me how important is each of the following in helping you to make decisions about your life.
Please indicate your answer by stating where you would place your feelings on a scale from 1 to 5 as shown on this hand Card.
D. Your own personal judgment
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{5} not very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1088/vshow}{GSS Data Explorer page for \code{decself}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab 2 \tab 3 \tab 4 \tab not very important \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 1052 \tab 250 \tab 111 \tab 31 \tab 26 \tab 4 \tab 7 \tab 1481 \cr
   Total \tab 0 \tab 1052 \tab 250 \tab 111 \tab 31 \tab 26 \tab 4 \tab 7 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{blkwhite}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{gochurch}},
\code{\link{permoral}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
