% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{decsorgs}
\alias{decsorgs}
\title{America should follow decision of intl org}
\source{
General Social Survey https://gss.norc.org
}
\description{
decsorgs
}
\details{
Question How much do you agree or disagree with the following statements? (Please, select one response)
C. In general, America should follow the decisions of international organizations to which it belongs, even if the government does not agree with them.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4885/vshow}{GSS Data Explorer page for \code{decsorgs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 53 \tab 286 \tab 392 \tab 342 \tab 65 \tab 77 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 37 \tab 330 \tab 355 \tab 346 \tab 76 \tab 125 \tab 5 \tab 2538 \cr
   Total \tab 2860 \tab 90 \tab 616 \tab 747 \tab 688 \tab 141 \tab 202 \tab 6 \tab 5350 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{citworld}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
