% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{dectreat}
\alias{dectreat}
\title{How able is x to decide receive treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
dectreat
}
\details{
Question In your opinion, how able is NAME to make his/her own decisions about the treatment s/he should receive -- very able, somewhat able, not very able or not able at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very able
\item \verb{[2]} somewhat able
\item \verb{[3]} not very able
\item \verb{[4]} not able at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1999/vshow}{GSS Data Explorer page for \code{dectreat}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very able \tab somewhat able \tab not very able \tab not able at all \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1460 \tab 294 \tab 422 \tab 445 \tab 241 \tab 28 \tab 14 \tab - \tab - \tab 2904 \cr
   2006 \tab 3073 \tab 335 \tab 519 \tab 364 \tab 163 \tab 40 \tab 16 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 263 \tab 425 \tab 314 \tab 125 \tab 17 \tab 29 \tab - \tab - \tab 2348 \cr
   2024 \tab 1691 \tab 343 \tab 630 \tab 451 \tab 140 \tab 23 \tab 10 \tab - \tab 21 \tab 3309 \cr
   Total \tab 7399 \tab 1235 \tab 1996 \tab 1574 \tab 669 \tab 108 \tab 69 \tab 0 \tab 21 \tab 13071 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}

Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
