% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_84.R
\name{defspdr}
\alias{defspdr}
\title{Defense spending--r's position}
\source{
General Social Survey https://gss.norc.org
}
\description{
defspdr
}
\details{
Question Some people believe that we should spend much less money for defense. Suppose these people are at one end of the scale at point number 1. Others feel that defense spending should be greatly increased. Suppose these people are at the other end, at point 7. And, of course, some other people have opinions somewhere in between at points 2, 3, 4, 5, or 6. ENTER CODE IN BOXES.
A. Where would you place yourself on this scale, or haven't you thought much about this?
}
\section{Values}{

\itemize{
\item \verb{[0]} havent thought much
\item \verb{[1]} greatly decrease
\item \verb{[7]} greatly increase
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/759/vshow}{GSS Data Explorer page for \code{defspdr}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab havent thought much \tab greatly decrease \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab greatly increase \tab don't know \tab no answer \tab Total \cr
   1980 \tab - \tab 135 \tab 47 \tab 49 \tab 97 \tab 308 \tab 335 \tab 189 \tab 278 \tab 25 \tab 5 \tab 1468 \cr
   Total \tab 0 \tab 135 \tab 47 \tab 49 \tab 97 \tab 308 \tab 335 \tab 189 \tab 278 \tab 25 \tab 5 \tab 1468 \cr
}
}

\seealso{
Other Core: 
\code{\link{defspdfg}},
\code{\link{sibdeath}},
\code{\link{spdeath}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}
}
\concept{Core}
\keyword{variable}
