% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{defund}
\alias{defund}
\title{Move funds for police to social service}
\source{
General Social Survey https://gss.norc.org
}
\description{
defund
}
\details{
Question Favor or oppose reducing funding for police and instead funding social services
}
\section{Values}{

\itemize{
\item \verb{[1]} favor
\item \verb{[2]} oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7574/vshow}{GSS Data Explorer page for \code{defund}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab favor \tab oppose \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1315 \tab 952 \tab 1392 \tab 7 \tab 332 \tab 34 \tab 4032 \cr
   Total \tab 1315 \tab 952 \tab 1392 \tab 7 \tab 332 \tab 34 \tab 4032 \cr
}
}

\keyword{variable}
