% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_92.R
\name{defwrknw}
\alias{defwrknw}
\title{Currently work for military supplier?}
\source{
General Social Survey https://gss.norc.org
}
\description{
defwrknw
}
\details{
Question Have you ever worked for a company where a major part of their business was selling supplies or services to the armed forces?
A. Do you work for such a company now?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/830/vshow}{GSS Data Explorer page for \code{defwrknw}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1984 \tab 1262 \tab 37 \tab 142 \tab 32 \tab 1473 \cr
   Total \tab 1262 \tab 37 \tab 142 \tab 32 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{resdefwk}}
}
\concept{Core}
\keyword{variable}
