% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{dejavu}
\alias{dejavu}
\title{Thought one was someplace before}
\source{
General Social Survey https://gss.norc.org
}
\description{
dejavu
}
\details{
Question How often have you had any of the following experiences?  READ EACH ITEM AND CIRCLE ONE CODE FOR EACH.
A. Thought you were somewhere you had been before, but knew that it was impossible.
}
\section{Values}{

\itemize{
\item \code{1} never in my life
\item \code{2} once or twice
\item \code{3} several times
\item \code{4} often
\item \code{8} cannot answer
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/316/vshow}{GSS Data Explorer page for \code{dejavu}}.
}

\seealso{
Other Core: 
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}},
\code{\link{visions}}

Other Parapsychological Experiences: 
\code{\link{esp}},
\code{\link{grace}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\concept{Parapsychological Experiences}
\keyword{variable}
