% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{dem10pst}
\alias{dem10pst}
\title{How well did democracy work in america ten yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
dem10pst
}
\details{
Question On the whole, on a scale of  0 to 10 where 0 is very poorly and 10 is very well.
B. And what about 10 years ago? How well did democracy work in America then?
}
\section{Values}{

\itemize{
\item \verb{[0]} very poorly
\item \verb{[10]} very well
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4961/vshow}{GSS Data Explorer page for \code{dem10pst}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab very poorly \tab 1 \tab very well \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   2004 \tab 1340 \tab 41 \tab 7 \tab 167 \tab 29 \tab 36 \tab 72 \tab 195 \tab 145 \tab 242 \tab 310 \tab 134 \tab 87 \tab 7 \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 24 \tab 7 \tab 75 \tab 33 \tab 55 \tab 84 \tab 223 \tab 121 \tab 218 \tab 203 \tab 75 \tab 96 \tab 50 \tab - \tab 2538 \cr
   2024 \tab 1663 \tab 50 \tab 15 \tab 143 \tab 37 \tab 61 \tab 103 \tab 217 \tab 175 \tab 260 \tab 248 \tab 106 \tab 203 \tab 28 \tab - \tab 3309 \cr
   Total \tab 4277 \tab 115 \tab 29 \tab 385 \tab 99 \tab 152 \tab 259 \tab 635 \tab 441 \tab 720 \tab 761 \tab 315 \tab 386 \tab 85 \tab 0 \tab 8659 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}

Other United States: 
\code{\link{dem10fut}},
\code{\link{demrghts}},
\code{\link{demtoday}}
}
\concept{ISSP Citizenship}
\concept{United States}
\keyword{variable}
