% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{demands}
\alias{demands}
\title{R feels that family make too many demands on r}
\source{
General Social Survey https://gss.norc.org
}
\description{
demands
}
\details{
Question Do you feel that your family, relatives and/or friends make too many demands on you?
}
\section{Values}{

\itemize{
\item \code{1} no, never
\item \code{2} yes, but seldom
\item \code{3} yes, sometimes
\item \code{4} yes, often
\item \code{5} yes, very often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4235/vshow}{GSS Data Explorer page for \code{demands}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{helpfrds}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
