% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{denkid}
\alias{denkid}
\title{If protestant, denomination r raised in}
\source{
General Social Survey https://gss.norc.org
}
\description{
denkid
}
\details{
Question What religion, if any, were you raised in? Was it Protestant, Catholic, Jewish, some other religion, or no religion?
A. If Protestant, what specific denomination was that?
}
\section{Values}{

\itemize{
\item \verb{[10000]} american baptist association
\item \verb{[11000]} american baptist churches in the u.s.a.
\item \verb{[12000]} national baptist convention of america
\item \verb{[13000]} national baptist convention u.s.a., inc.
\item \verb{[14000]} southern baptist convention
\item \verb{[15000]} other baptist churches
\item \verb{[18000]} baptist, don't know which
\item \verb{[20000]} african methodist episcopal church
\item \verb{[21000]} african methodist episcopal zion church
\item \verb{[22000]} united methodist church
\item \verb{[23000]} other methodist churches
\item \verb{[28000]} methodist, don't know which
\item \verb{[30000]} american lutheran church
\item \verb{[31000]} lutheran church in america
\item \verb{[32000]} lutheran church - missouri synod
\item \verb{[33000]} wisconsin evangelical lutheran synod
\item \verb{[34000]} other lutheran churches
\item \verb{[35000]} evangelical lutheran
\item \verb{[38000]} lutheran, don't know which
\item \verb{[40000]} presbyterian church in the u.s.
\item \verb{[41000]} united presbyterian church in the u.s.a.
\item \verb{[42000]} other presbyterian churches
\item \verb{[43000]} presbyterian church u.s.a.
\item \verb{[48000]} presbyterian, don't know which
\item \verb{[50000]} episcopal church
\item \verb{[60001]} hungarian reformed
\item \verb{[60002]} evangelical congregational
\item \verb{[60003]} ind bible, bible, bible fellowship
\item \verb{[60005]} church of prophecy
\item \verb{[60006]} new testament christian
\item \verb{[60007]} church of god, saint & christ
\item \verb{[60008]} moravian
\item \verb{[60009]} christian & missionary alliance
\item \verb{[60010]} advent christian
\item \verb{[60011]} spiritualist
\item \verb{[60012]} assembly of god
\item \verb{[60013]} free methodist
\item \verb{[60014]} apostolic faith
\item \verb{[60015]} african methodist
\item \verb{[60016]} free will baptist
\item \verb{[60017]} eden evangelist
\item \verb{[60018]} holiness (nazarene)
\item \verb{[60019]} baptist (northern)
\item \verb{[60020]} brethren church, brethren
\item \verb{[60021]} witness holiness
\item \verb{[60022]} brethren, plymouth
\item \verb{[60023]} united brethren, united brethren in christ
\item \verb{[60024]} independent
\item \verb{[60025]} christian disciples
\item \verb{[60026]} christ in christian union
\item \verb{[60027]} open bible
\item \verb{[60028]} christian catholic
\item \verb{[60029]} christ church unity
\item \verb{[60030]} christ adelphians
\item \verb{[60031]} christian; central christian
\item \verb{[60032]} christian reform
\item \verb{[60033]} christian scientist
\item \verb{[60034]} church of christ, evangelical
\item \verb{[60035]} church of christ
\item \verb{[60036]} churches of god (except with christ and holiness)
\item \verb{[60037]} church of god in christ
\item \verb{[60038]} church of god in christ holiness
\item \verb{[60039]} church of the living god
\item \verb{[60040]} congregationalist, 1st congreg
\item \verb{[60041]} community church
\item \verb{[60042]} covenant
\item \verb{[60043]} dutch reform
\item \verb{[60044]} disciples of christ
\item \verb{[60045]} evangelical, evangelist
\item \verb{[60046]} evangelical reformed
\item \verb{[60047]} evangelist free church
\item \verb{[60048]} first church
\item \verb{[60049]} first christian disciples of christ
\item \verb{[60050]} first reformed
\item \verb{[60051]} first christian
\item \verb{[60052]} full gospel
\item \verb{[60053]} four square gospel
\item \verb{[60054]} friends
\item \verb{[60055]} holy roller
\item \verb{[60056]} holiness; church of holiness
\item \verb{[60057]} pilgrim holiness
\item \verb{[60058]} jehovah''s witnesses
\item \verb{[60059]} lds
\item \verb{[60060]} lds--mormon
\item \verb{[60061]} lds--reorganized
\item \verb{[60062]} lds--jesus christ; church of jesus lds
\item \verb{[60063]} mennonite
\item \verb{[60064]} mormon
\item \verb{[60065]} nazarene
\item \verb{[60066]} pentecostal assembly of god
\item \verb{[60067]} pentecostal church of god
\item \verb{[60068]} pentecostal
\item \verb{[60069]} pentecostal holiness, holiness pentecostal
\item \verb{[60070]} quaker
\item \verb{[60071]} reformed
\item \verb{[60072]} reformed united church of christ
\item \verb{[60073]} reformed church of christ
\item \verb{[60074]} religious science
\item \verb{[60075]} mind science
\item \verb{[60076]} salvation army
\item \verb{[60077]} 7th day adventist
\item \verb{[60078]} sanctified, sanctification
\item \verb{[60079]} united holiness
\item \verb{[60080]} unitarian, universalist
\item \verb{[60081]} united church of christ
\item \verb{[60082]} united church, unity church
\item \verb{[60083]} wesleyan
\item \verb{[60084]} wesleyan methodist--pilgrim
\item \verb{[60085]} zion union
\item \verb{[60086]} zion union apostolic
\item \verb{[60087]} zion union apostolic--reformed
\item \verb{[60088]} disciples of god
\item \verb{[60089]} grace reformed
\item \verb{[60090]} holiness church of god
\item \verb{[60091]} evangelical covenant
\item \verb{[60092]} mission covenant
\item \verb{[60093]} missionary baptist
\item \verb{[60094]} swedish mission
\item \verb{[60095]} unity
\item \verb{[60096]} united church of christianity
\item \verb{[60097]} other fundamentalist
\item \verb{[60098]} federated church
\item \verb{[60099]} american reform
\item \verb{[60100]} grace brethren
\item \verb{[60101]} christ in god
\item \verb{[60102]} charismatic
\item \verb{[60103]} pentecostal apostolic
\item \verb{[60104]} house of prayer
\item \verb{[60105]} latvian lutheran
\item \verb{[60106]} triumph church of god
\item \verb{[60107]} apostolic christian
\item \verb{[60108]} christ cathedral of truth
\item \verb{[60109]} bible missionary
\item \verb{[60110]} calvary bible
\item \verb{[60111]} amish
\item \verb{[60112]} evangelical methodist
\item \verb{[60113]} worldwide church of god
\item \verb{[60114]} church universal and triumphant
\item \verb{[60115]} mennonite brethren
\item \verb{[60116]} church of the first born
\item \verb{[60117]} missionary church
\item \verb{[60118]} the way ministry
\item \verb{[60119]} united church of canada
\item \verb{[60120]} evangelical united brethren
\item \verb{[60121]} the church of god of prophecy
\item \verb{[60122]} chapel of faith
\item \verb{[60123]} polish national church
\item \verb{[60124]} faith gospel tabernacle
\item \verb{[60125]} christian calvary chapel
\item \verb{[60126]} carmelite
\item \verb{[60127]} church of daniel's band
\item \verb{[60128]} christian tabernacle
\item \verb{[60129]} living word
\item \verb{[60130]} true light church of christ
\item \verb{[60131]} macedonia
\item \verb{[60132]} brother of christ
\item \verb{[60133]} primitive baptist
\item \verb{[60134]} independent fundamental church of america
\item \verb{[60135]} chinese gospel church
\item \verb{[60136]} new age spirituality
\item \verb{[60137]} new song
\item \verb{[60138]} apostolic church
\item \verb{[60139]} faith christian
\item \verb{[60140]} people's church
\item \verb{[60141]} new birth christian
\item \verb{[60142]} unity school of christianity
\item \verb{[60143]} assyrian evangelist church
\item \verb{[60144]} spirit of christ
\item \verb{[60145]} church of jesus christ of the restoration
\item \verb{[60146]} laotian christian
\item \verb{[60148]} schwenkfelder
\item \verb{[60149]} polish catholic
\item \verb{[60150]} zwinglian
\item \verb{[60151]} world overcomer outreach ministry
\item \verb{[60152]} course in miracles
\item \verb{[60153]} unity of the brethren
\item \verb{[60154]} spirit filled
\item \verb{[60155]} christian union
\item \verb{[60156]} church of living christ
\item \verb{[60157]} community of christ
\item \verb{[60158]} new hope christian fellowship
\item \verb{[60159]} community christian fellowship
\item \verb{[60160]} friends in christ
\item \verb{[60161]} hawaiian ohana
\item \verb{[60162]} reformed church of jesus christ of latter day saints
\item \verb{[60163]} swedenurgian/churches of the new jerusalem
\item \verb{[60164]} divine science
\item \verb{[60166]} united christian
\item \verb{[60167]} sanctuary
\item \verb{[60168]} rain on us deliverance ministries
\item \verb{[60169]} the word church
\item \verb{[60170]} cornerstone church
\item \verb{[60171]} life sanctuary
\item \verb{[60172]} word of faith church
\item \verb{[60173]} harvest church
\item \verb{[60174]} shephard''s chapel
\item \verb{[60175]} greater new testament church
\item \verb{[60176]} vineyard church
\item \verb{[60177]} real life ministries
\item \verb{[60178]} cathedral of joy
\item \verb{[60179]} great faith ministries
\item \verb{[60180]} shield of faith ministries
\item \verb{[60181]} born again
\item \verb{[60182]} alliance
\item \verb{[60183]} jacobite apostolic
\item \verb{[60184]} church of god of israel
\item \verb{[60185]} journeys
\item \verb{[60186]} national progressive baptist
\item \verb{[60187]} new apostolic
\item \verb{[60188]} metropolitan community
\item \verb{[60189]} family life church
\item \verb{[60190]} faith fellowship
\item \verb{[60191]} faith covenant
\item \verb{[60192]} new thought
\item \verb{[60193]} free spirit ministry
\item \verb{[60194]} jacobite syrian christian church
\item \verb{[60195]} the ark church
\item \verb{[60196]} empowerment temple
\item \verb{[60197]} grace independent baptist church
\item \verb{[60198]} new life
\item \verb{[60199]} follower of christ
\item \verb{[60200]} north country chapel
\item \verb{[60201]} pathways christian church
\item \verb{[60202]} messianic christian
\item \verb{[60203]} warehouse ministries
\item \verb{[60204]} universal church of the kingdom of god
\item \verb{[60205]} assembly of christ
\item \verb{[60206]} the amana church
\item \verb{[60207]} the legacy church
\item \verb{[60208]} calvary
\item \verb{[60209]} church of the living god
\item \verb{[60210]} ethiopian evangelical church
\item \verb{[60211]} disciple of jesus
\item \verb{[60212]} scandinavian church
\item \verb{[60213]} hebrew roots
\item \verb{[60214]} hebrew isrealites
\item \verb{[60215]} armenian apostolic church
\item \verb{[60216]} word of jesus
\item \verb{[60217]} old testament
\item \verb{[70000]} no denomination given or non-denom. church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4670/vshow}{GSS Data Explorer page for \code{denkid}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab american baptist association \tab american baptist churches in the u.s.a. \tab national baptist convention of america \tab southern baptist convention \tab other baptist churches \tab baptist, don't know which \tab african methodist episcopal church \tab african methodist episcopal zion church \tab united methodist church \tab other methodist churches \tab methodist, don't know which \tab american lutheran church \tab lutheran church in america \tab lutheran church - missouri synod \tab wisconsin evangelical lutheran synod \tab other lutheran churches \tab evangelical lutheran \tab lutheran, don't know which \tab presbyterian church in the u.s. \tab united presbyterian church in the u.s.a. \tab other presbyterian churches \tab presbyterian church u.s.a. \tab presbyterian, don't know which \tab episcopal church \tab moravian \tab assembly of god \tab free methodist \tab african methodist \tab free will baptist \tab united brethren, united brethren in christ \tab christian; central christian \tab christian reform \tab christian scientist \tab church of christ \tab churches of god (except with christ and holiness) \tab church of god in christ \tab congregationalist, 1st congreg \tab covenant \tab dutch reform \tab disciples of christ \tab evangelical, evangelist \tab evangelist free church \tab first christian disciples of christ \tab first reformed \tab first christian \tab holiness; church of holiness \tab jehovah''s witnesses \tab lds \tab lds--mormon \tab lds--reorganized \tab lds--jesus christ; church of jesus lds \tab mennonite \tab mormon \tab nazarene \tab pentecostal assembly of god \tab pentecostal \tab quaker \tab reformed \tab 7th day adventist \tab sanctified, sanctification \tab unitarian, universalist \tab united church of christ \tab wesleyan \tab wesleyan methodist--pilgrim \tab mission covenant \tab missionary baptist \tab evangelical methodist \tab worldwide church of god \tab united church of canada \tab evangelical united brethren \tab new birth christian \tab no denomination given or non-denom. church \tab don't know \tab no answer \tab brethren church, brethren \tab independent \tab church of god in christ holiness \tab full gospel \tab apostolic church \tab polish catholic \tab national baptist convention u.s.a., inc. \tab ind bible, bible, bible fellowship \tab christian & missionary alliance \tab apostolic faith \tab holiness (nazarene) \tab community church \tab pilgrim holiness \tab charismatic \tab missionary church \tab spiritualist \tab evangelical reformed \tab armenian apostolic church \tab Total \cr
   1991 \tab 704 \tab 22 \tab 7 \tab 5 \tab 94 \tab 9 \tab 92 \tab 11 \tab 3 \tab 87 \tab 2 \tab 27 \tab 18 \tab 6 \tab 22 \tab 3 \tab 7 \tab 5 \tab 21 \tab 10 \tab 21 \tab 3 \tab 5 \tab 16 \tab 24 \tab 1 \tab 10 \tab 2 \tab 1 \tab 3 \tab 1 \tab 17 \tab 2 \tab 4 \tab 14 \tab 9 \tab 2 \tab 6 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 7 \tab 3 \tab 17 \tab 2 \tab 1 \tab 2 \tab 4 \tab 3 \tab 4 \tab 1 \tab 14 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 6 \tab 100 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1998 \tab 2083 \tab 1 \tab - \tab - \tab 42 \tab 4 \tab 165 \tab 28 \tab 1 \tab 25 \tab 2 \tab 79 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 4 \tab 47 \tab - \tab 1 \tab - \tab - \tab 39 \tab 9 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 17 \tab 1 \tab 3 \tab 15 \tab 3 \tab 1 \tab 9 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 7 \tab 4 \tab - \tab 9 \tab - \tab 1 \tab 12 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 10 \tab 11 \tab 147 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2008 \tab 1269 \tab 16 \tab 7 \tab 7 \tab 108 \tab 1 \tab 138 \tab 7 \tab 3 \tab 66 \tab 1 \tab 24 \tab 11 \tab 4 \tab 13 \tab 5 \tab - \tab 2 \tab 23 \tab 8 \tab 6 \tab - \tab 2 \tab 30 \tab 31 \tab 1 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab 33 \tab 3 \tab - \tab 11 \tab 6 \tab 1 \tab 6 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 5 \tab 10 \tab 10 \tab 4 \tab - \tab - \tab 1 \tab 10 \tab 4 \tab - \tab 20 \tab 4 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 40 \tab 11 \tab 13 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2023 \cr
   2018 \tab 1752 \tab - \tab 2 \tab - \tab 17 \tab 3 \tab 193 \tab 1 \tab 1 \tab 13 \tab 1 \tab 53 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 49 \tab - \tab - \tab - \tab 1 \tab 25 \tab 18 \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab 51 \tab - \tab - \tab 11 \tab 4 \tab - \tab 7 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 4 \tab 11 \tab - \tab - \tab - \tab - \tab 2 \tab 14 \tab 4 \tab - \tab 18 \tab 3 \tab 2 \tab 7 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 25 \tab 17 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2348 \cr
   Total \tab 5808 \tab 39 \tab 16 \tab 12 \tab 261 \tab 17 \tab 588 \tab 47 \tab 8 \tab 191 \tab 6 \tab 183 \tab 32 \tab 11 \tab 43 \tab 11 \tab 8 \tab 12 \tab 140 \tab 18 \tab 28 \tab 3 \tab 8 \tab 110 \tab 82 \tab 3 \tab 20 \tab 2 \tab 1 \tab 12 \tab 2 \tab 118 \tab 6 \tab 7 \tab 51 \tab 22 \tab 4 \tab 28 \tab 3 \tab 2 \tab 13 \tab 8 \tab 1 \tab 1 \tab 1 \tab 7 \tab 19 \tab 31 \tab 27 \tab 6 \tab 1 \tab 2 \tab 7 \tab 34 \tab 16 \tab 1 \tab 61 \tab 8 \tab 7 \tab 27 \tab 1 \tab 6 \tab 10 \tab 4 \tab 2 \tab 1 \tab 13 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 88 \tab 45 \tab 261 \tab 6 \tab 4 \tab 3 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{relexper}},
\code{\link{religkid}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\keyword{variable}
