% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{denom}
\alias{denom}
\title{Specific denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
denom
}
\details{
Question What is your religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?  A. What specific denomination is that, if any?
As of 2024, DENOM has been recoded to only account for seven broad denomination categories. This has been retroactively applied to all previous years of DENOM. The original DENOM containing data 1972 to 2022 has been renamed DENOM_7222.
}
\section{Values}{

\itemize{
\item \verb{[1]} baptist
\item \verb{[2]} methodist
\item \verb{[3]} lutheran
\item \verb{[4]} presbyterian
\item \verb{[5]} episcopal
\item \verb{[6]} other denomination
\item \verb{[7]} no denomination or non-denominational church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/288/vshow}{GSS Data Explorer page for \code{denom}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab baptist \tab methodist \tab lutheran \tab presbyterian \tab episcopal \tab other denomination \tab no denomination or non-denominational church \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 324 \tab 232 \tab 139 \tab 80 \tab 33 \tab 185 \tab 36 \tab 577 \tab 7 \tab - \tab - \tab 1613 \cr
   1973 \tab 312 \tab 198 \tab 124 \tab 57 \tab 41 \tab 163 \tab 43 \tab 564 \tab 2 \tab - \tab - \tab 1504 \cr
   1974 \tab 322 \tab 190 \tab 113 \tab 74 \tab 38 \tab 174 \tab 42 \tab 529 \tab 2 \tab - \tab - \tab 1484 \cr
   1975 \tab 309 \tab 175 \tab 139 \tab 76 \tab 44 \tab 182 \tab 49 \tab 513 \tab 3 \tab - \tab - \tab 1490 \cr
   1976 \tab 303 \tab 162 \tab 104 \tab 78 \tab 49 \tab 195 \tab 60 \tab 546 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab 324 \tab 193 \tab 136 \tab 66 \tab 37 \tab 190 \tab 58 \tab 519 \tab 7 \tab - \tab - \tab 1530 \cr
   1978 \tab 318 \tab 190 \tab 115 \tab 62 \tab 39 \tab 208 \tab 46 \tab 549 \tab 5 \tab - \tab - \tab 1532 \cr
   1980 \tab 307 \tab 170 \tab 92 \tab 79 \tab 41 \tab 195 \tab 52 \tab 529 \tab 3 \tab - \tab - \tab 1468 \cr
   1982 \tab 478 \tab 210 \tab 126 \tab 75 \tab 52 \tab 244 \tab 72 \tab 590 \tab 13 \tab - \tab - \tab 1860 \cr
   1983 \tab 303 \tab 147 \tab 143 \tab 64 \tab 37 \tab 229 \tab 45 \tab 625 \tab 6 \tab - \tab - \tab 1599 \cr
   1984 \tab 295 \tab 164 \tab 93 \tab 75 \tab 28 \tab 220 \tab 57 \tab 529 \tab 12 \tab - \tab - \tab 1473 \cr
   1985 \tab 324 \tab 169 \tab 99 \tab 57 \tab 42 \tab 215 \tab 48 \tab 574 \tab 6 \tab - \tab - \tab 1534 \cr
   1986 \tab 295 \tab 145 \tab 115 \tab 50 \tab 38 \tab 224 \tab 53 \tab 546 \tab 4 \tab - \tab - \tab 1470 \cr
   1987 \tab 517 \tab 172 \tab 86 \tab 80 \tab 36 \tab 299 \tab 61 \tab 558 \tab 10 \tab - \tab - \tab 1819 \cr
   1988 \tab 314 \tab 139 \tab 74 \tab 63 \tab 31 \tab 216 \tab 65 \tab 574 \tab 4 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 305 \tab 148 \tab 112 \tab 70 \tab 43 \tab 204 \tab 86 \tab 563 \tab 6 \tab - \tab - \tab 1537 \cr
   1990 \tab 231 \tab 129 \tab 107 \tab 69 \tab 34 \tab 211 \tab 78 \tab 505 \tab 7 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 305 \tab 137 \tab 104 \tab 74 \tab 32 \tab 260 \tab 53 \tab 549 \tab 3 \tab - \tab - \tab 1517 \cr
   1993 \tab 312 \tab 170 \tab 117 \tab 71 \tab 34 \tab 235 \tab 84 \tab 572 \tab 11 \tab - \tab - \tab 1606 \cr
   1994 \tab 616 \tab 280 \tab 199 \tab 142 \tab 66 \tab 338 \tab 125 \tab 1207 \tab 18 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 628 \tab 244 \tab 198 \tab 106 \tab 79 \tab 304 \tab 99 \tab 1236 \tab 10 \tab - \tab - \tab 2904 \cr
   1998 \tab 568 \tab 249 \tab 171 \tab 87 \tab 60 \tab 266 \tab 139 \tab 1245 \tab 46 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 557 \tab 252 \tab 145 \tab 92 \tab 52 \tab 287 \tab 161 \tab 1254 \tab 16 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 482 \tab 206 \tab 149 \tab 79 \tab 58 \tab 338 \tab 203 \tab 1225 \tab 17 \tab 8 \tab - \tab 2765 \cr
   2004 \tab 498 \tab 191 \tab 127 \tab 67 \tab 59 \tab 348 \tab 263 \tab 1239 \tab 16 \tab 4 \tab - \tab 2812 \cr
   2006 \tab 782 \tab 295 \tab 181 \tab 124 \tab 78 \tab 581 \tab 370 \tab 2054 \tab 28 \tab 17 \tab - \tab 4510 \cr
   2008 \tab 329 \tab 131 \tab 61 \tab 56 \tab 43 \tab 289 \tab 191 \tab 907 \tab 9 \tab 7 \tab - \tab 2023 \cr
   2010 \tab 355 \tab 114 \tab 82 \tab 60 \tab 36 \tab 209 \tab 198 \tab 969 \tab 13 \tab 8 \tab - \tab 2044 \cr
   2012 \tab 317 \tab 116 \tab 82 \tab 33 \tab 32 \tab 230 \tab 222 \tab 935 \tab 3 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 368 \tab 143 \tab 92 \tab 61 \tab 39 \tab 260 \tab 293 \tab 1264 \tab 15 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 441 \tab 129 \tab 111 \tab 53 \tab 40 \tab 301 \tab 336 \tab 1432 \tab 16 \tab 8 \tab - \tab 2867 \cr
   2018 \tab 336 \tab 135 \tab 85 \tab 56 \tab 26 \tab 248 \tab 266 \tab 1163 \tab 20 \tab 13 \tab - \tab 2348 \cr
   2021 \tab 409 \tab 191 \tab 161 \tab 103 \tab 55 \tab 219 \tab 551 \tab 2243 \tab 46 \tab 15 \tab 39 \tab 4032 \cr
   2022 \tab 409 \tab 169 \tab 108 \tab 59 \tab 44 \tab 206 \tab 498 \tab 2010 \tab 32 \tab - \tab 9 \tab 3544 \cr
   2024 \tab 429 \tab 125 \tab 120 \tab 65 \tab 42 \tab 212 \tab 406 \tab 1891 \tab 6 \tab 1 \tab 12 \tab 3309 \cr
   Total \tab 13722 \tab 6210 \tab 4210 \tab 2563 \tab 1538 \tab 8685 \tab 5409 \tab 32785 \tab 424 \tab 93 \tab 60 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}},
\code{\link{wirtap}}

Other Religion: 
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
