% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{denom16}
\alias{denom16}
\title{Denomination in which r was raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
denom16
}
\details{
Question In what religion were you raised?  A. What specific denomination is that, if any?
As of 2024, DENOM16 has been recoded to only account for seven broad denomination categories. This has been retroactively applied to all previous years of DENOM16. The original DENOM16 containing data 1973 to 2022 has been renamed DENOM16_7322.
}
\section{Values}{

\itemize{
\item \verb{[1]} baptist
\item \verb{[2]} methodist
\item \verb{[3]} lutheran
\item \verb{[4]} presbyterian
\item \verb{[5]} episcopal
\item \verb{[6]} other denomination
\item \verb{[7]} no denomination or non-denominational church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/342/vshow}{GSS Data Explorer page for \code{denom16}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab baptist \tab methodist \tab lutheran \tab presbyterian \tab episcopal \tab other denomination \tab no denomination or non-denominational church \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1973 \tab - \tab 354 \tab 198 \tab 131 \tab 73 \tab 41 \tab 151 \tab 26 \tab 523 \tab 7 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 366 \tab 217 \tab 119 \tab 81 \tab 31 \tab 146 \tab 22 \tab 500 \tab 2 \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 343 \tab 235 \tab 149 \tab 67 \tab 39 \tab 167 \tab 30 \tab 458 \tab 1 \tab 1 \tab - \tab 1490 \cr
   1976 \tab - \tab 340 \tab 196 \tab 107 \tab 95 \tab 38 \tab 167 \tab 27 \tab 526 \tab 3 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 367 \tab 221 \tab 145 \tab 78 \tab 36 \tab 163 \tab 31 \tab 482 \tab 7 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 371 \tab 222 \tab 127 \tab 63 \tab 29 \tab 171 \tab 32 \tab 508 \tab 9 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 330 \tab 224 \tab 105 \tab 91 \tab 37 \tab 168 \tab 25 \tab 483 \tab 5 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 550 \tab 272 \tab 128 \tab 81 \tab 49 \tab 181 \tab 33 \tab 556 \tab 10 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 340 \tab 183 \tab 143 \tab 62 \tab 40 \tab 201 \tab 26 \tab 594 \tab 10 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 327 \tab 197 \tab 92 \tab 79 \tab 26 \tab 189 \tab 36 \tab 499 \tab 28 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 348 \tab 201 \tab 96 \tab 72 \tab 39 \tab 199 \tab 23 \tab 551 \tab 5 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 346 \tab 182 \tab 119 \tab 47 \tab 35 \tab 193 \tab 26 \tab 515 \tab 5 \tab 2 \tab - \tab 1470 \cr
   1987 \tab - \tab 588 \tab 221 \tab 84 \tab 82 \tab 31 \tab 252 \tab 26 \tab 500 \tab 34 \tab 1 \tab - \tab 1819 \cr
   1988 \tab - \tab 363 \tab 184 \tab 88 \tab 82 \tab 31 \tab 171 \tab 30 \tab 529 \tab 3 \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 321 \tab 169 \tab 121 \tab 74 \tab 32 \tab 186 \tab 61 \tab 567 \tab 6 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 272 \tab 164 \tab 111 \tab 67 \tab 30 \tab 190 \tab 40 \tab 487 \tab 11 \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 343 \tab 179 \tab 109 \tab 71 \tab 31 \tab 193 \tab 37 \tab 548 \tab 6 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 332 \tab 185 \tab 116 \tab 86 \tab 41 \tab 237 \tab 37 \tab 557 \tab 15 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 706 \tab 319 \tab 220 \tab 139 \tab 68 \tab 309 \tab 69 \tab 1139 \tab 21 \tab 2 \tab - \tab 2992 \cr
   1996 \tab - \tab 666 \tab 338 \tab 193 \tab 120 \tab 73 \tab 271 \tab 60 \tab 1165 \tab 16 \tab 2 \tab - \tab 2904 \cr
   1998 \tab - \tab 635 \tab 317 \tab 179 \tab 124 \tab 68 \tab 251 \tab 60 \tab 1147 \tab 48 \tab 3 \tab - \tab 2832 \cr
   2000 \tab - \tab 645 \tab 298 \tab 159 \tab 105 \tab 47 \tab 272 \tab 91 \tab 1172 \tab 25 \tab 3 \tab - \tab 2817 \cr
   2002 \tab - \tab 567 \tab 240 \tab 175 \tab 104 \tab 70 \tab 317 \tab 107 \tab 1159 \tab 20 \tab 6 \tab - \tab 2765 \cr
   2004 \tab - \tab 574 \tab 256 \tab 144 \tab 89 \tab 57 \tab 329 \tab 152 \tab 1196 \tab 10 \tab 5 \tab - \tab 2812 \cr
   2006 \tab - \tab 605 \tab 239 \tab 154 \tab 104 \tab 61 \tab 356 \tab 100 \tab 2863 \tab 2 \tab 26 \tab - \tab 4510 \cr
   2008 \tab - \tab 416 \tab 155 \tab 76 \tab 69 \tab 40 \tab 269 \tab 86 \tab 891 \tab 12 \tab 9 \tab - \tab 2023 \cr
   2010 \tab - \tab 434 \tab 145 \tab 95 \tab 86 \tab 48 \tab 182 \tab 143 \tab 888 \tab 10 \tab 13 \tab - \tab 2044 \cr
   2012 \tab - \tab 399 \tab 156 \tab 99 \tab 48 \tab 38 \tab 198 \tab 122 \tab 905 \tab 4 \tab 5 \tab - \tab 1974 \cr
   2014 \tab - \tab 475 \tab 171 \tab 113 \tab 78 \tab 44 \tab 293 \tab 156 \tab 1185 \tab 12 \tab 11 \tab - \tab 2538 \cr
   2016 \tab - \tab 561 \tab 200 \tab 134 \tab 75 \tab 54 \tab 302 \tab 191 \tab 1316 \tab 11 \tab 23 \tab - \tab 2867 \cr
   2018 \tab - \tab 477 \tab 155 \tab 107 \tab 62 \tab 37 \tab 237 \tab 151 \tab 1090 \tab 12 \tab 20 \tab - \tab 2348 \cr
   2021 \tab - \tab 566 \tab 309 \tab 208 \tab 147 \tab 74 \tab 291 \tab 289 \tab 2108 \tab 4 \tab 29 \tab 7 \tab 4032 \cr
   2022 \tab - \tab 569 \tab 201 \tab 145 \tab 95 \tab 63 \tab 281 \tab 324 \tab 1749 \tab 18 \tab 78 \tab 21 \tab 3544 \cr
   2024 \tab - \tab 559 \tab 199 \tab 174 \tab 92 \tab 56 \tab 250 \tab 208 \tab 1703 \tab 12 \tab 34 \tab 22 \tab 3309 \cr
   Total \tab 0 \tab 15455 \tab 7348 \tab 4465 \tab 2888 \tab 1534 \tab 7733 \tab 2877 \tab 31059 \tab 404 \tab 273 \tab 50 \tab 74086 \cr
}
}

\seealso{
Other Core: 
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}

Other Religion: 
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
