% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{denom_next}
\alias{denom_next}
\title{Denomination - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
denom_next
}
\details{
Question What specific denomination is that, if any?
}
\section{Values}{

\itemize{
\item \verb{[10]} american baptist association
\item \verb{[11]} american baptist churches in the u.s.a.
\item \verb{[12]} national baptist convention of america
\item \verb{[13]} national baptist convention u.s.a., inc.
\item \verb{[14]} southern baptist convention
\item \verb{[15]} other baptist churches
\item \verb{[18]} baptist, don't know which
\item \verb{[20]} african methodist episcopal church
\item \verb{[21]} african methodist episcopal zion church
\item \verb{[22]} united methodist church
\item \verb{[23]} other methodist churches
\item \verb{[28]} methodist, don't know which
\item \verb{[30]} american lutheran church
\item \verb{[31]} lutheran church in america
\item \verb{[32]} lutheran church - missouri synod
\item \verb{[33]} wisconsin evangelical lutheran synod
\item \verb{[34]} other lutheran churches
\item \verb{[35]} evangelical lutheran
\item \verb{[38]} lutheran, don't know which
\item \verb{[40]} presbyterian church in the u.s.
\item \verb{[41]} united presbyterian church in the u.s.a.
\item \verb{[42]} other presbyterian churches
\item \verb{[43]} presbyterian, merged
\item \verb{[48]} presbyterian, don't know which
\item \verb{[50]} episcopal church
\item \verb{[60]} other, specify
\item \verb{[70]} no denomination or non-denominational church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8002/vshow}{GSS Data Explorer page for \code{denom_next}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab iap \tab american baptist association \tab american baptist churches in the u.s.a. \tab southern baptist convention \tab other baptist churches \tab baptist, don't know which \tab african methodist episcopal church \tab african methodist episcopal zion church \tab united methodist church \tab other methodist churches \tab methodist, don't know which \tab american lutheran church \tab lutheran church in america \tab lutheran church - missouri synod \tab other lutheran churches \tab evangelical lutheran \tab lutheran, don't know which \tab presbyterian church in the u.s. \tab united presbyterian church in the u.s.a. \tab other presbyterian churches \tab presbyterian, merged \tab presbyterian, don't know which \tab episcopal church \tab other, specify \tab no denomination or non-denominational church \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 852 \tab 8 \tab 1 \tab 30 \tab 24 \tab 22 \tab 1 \tab 1 \tab 41 \tab 2 \tab 4 \tab 4 \tab 1 \tab 14 \tab 3 \tab 11 \tab 8 \tab 5 \tab 4 \tab 2 \tab 6 \tab 5 \tab 14 \tab 52 \tab 116 \tab 2308 \tab 5 \tab 3544 \cr
   Total \tab 852 \tab 8 \tab 1 \tab 30 \tab 24 \tab 22 \tab 1 \tab 1 \tab 41 \tab 2 \tab 4 \tab 4 \tab 1 \tab 14 \tab 3 \tab 11 \tab 8 \tab 5 \tab 4 \tab 2 \tab 6 \tab 5 \tab 14 \tab 52 \tab 116 \tab 2308 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
