% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{denomsch}
\alias{denomsch}
\title{Denomination of school children attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
denomsch
}
\details{
Question Do/did all of your children attend public school for all of their elementary and high school education?
A. Which types of school do/did your children attend?
3a. What denomination or church was/is this school affiliated with?
}
\section{Values}{

\itemize{
\item \verb{[1]} baptist
\item \verb{[2]} christian reform
\item \verb{[3]} christian science
\item \verb{[4]} episcopal
\item \verb{[5]} inter denominational
\item \verb{[6]} lutheran
\item \verb{[7]} missouri synod lutheran
\item \verb{[8]} non-denominational
\item \verb{[9]} southern baptist
\item \verb{[10]} seventh day adventist
\item \verb{[11]} valley christian
\item \verb{[12]} congregational
\item \verb{[13]} mormon
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1151/vshow}{GSS Data Explorer page for \code{denomsch}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab baptist \tab christian reform \tab christian science \tab episcopal \tab lutheran \tab missouri synod lutheran \tab non-denominational \tab southern baptist \tab seventh day adventist \tab valley christian \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2739 \tab 13 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 44 \tab 2817 \cr
   Total \tab 2739 \tab 13 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 44 \tab 2817 \cr
}
}

\seealso{
Other Religion: 
\code{\link{chrissch}},
\code{\link{implives}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}}
}
\concept{Religion}
\keyword{variable}
