% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{densp}
\alias{densp}
\title{If protestant, r's spouse's denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
densp
}
\details{
Question What is your husband's/wife's religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
A. If Protestant, what specific denomination was that?
}
\section{Values}{

\itemize{
\item \verb{[10000]} amer baptist assoc
\item \verb{[11000]} amer baptist ch in usa
\item \verb{[12000]} natl baptist conv. of america
\item \verb{[13000]} natl baptist conv., usa, inc
\item \verb{[14000]} southern baptist conv
\item \verb{[15000]} other baptist chs
\item \verb{[18000]} baptist, dk which
\item \verb{[20000]} african meth epis ch
\item \verb{[21000]} ame zion church
\item \verb{[22000]} united meth church
\item \verb{[23000]} other meth ch
\item \verb{[28000]} methodist, dk which
\item \verb{[30000]} amer lutheran ch
\item \verb{[31000]} luth ch in amer
\item \verb{[32000]} luth ch--mo synod
\item \verb{[33000]} wisc evan luth synod
\item \verb{[34000]} other luth chs
\item \verb{[35000]} evan lutheran
\item \verb{[38000]} lutheran, dk which
\item \verb{[40000]} presbyt ch in usa
\item \verb{[41000]} united presbyt ch in usa
\item \verb{[42000]} other presbyt chs
\item \verb{[43000]} presbyt ch usa
\item \verb{[48000]} presbyterian, dk which
\item \verb{[50000]} episcopal church
\item \verb{[60002]} evangeligal congregational
\item \verb{[60003]} ind. bible fellowship
\item \verb{[60006]} new testament christian
\item \verb{[60008]} moravian
\item \verb{[60009]} christ. and mission. alliance
\item \verb{[60010]} adventist
\item \verb{[60012]} assembly of god
\item \verb{[60013]} free methodist
\item \verb{[60014]} apostolic
\item \verb{[60015]} african methodist
\item \verb{[60016]} free will baptist
\item \verb{[60018]} holiness, nazarene
\item \verb{[60020]} brethren church
\item \verb{[60023]} united brethren, breth. in christ
\item \verb{[60024]} independent
\item \verb{[60031]} christian; central christian
\item \verb{[60032]} christian reform
\item \verb{[60033]} christian scientist
\item \verb{[60035]} church of christ
\item \verb{[60036]} churches of god -exc. christ & holiness
\item \verb{[60037]} church of god in christ
\item \verb{[60038]} holiness
\item \verb{[60040]} congregationalist, 1st cong
\item \verb{[60041]} community church
\item \verb{[60042]} covenant
\item \verb{[60043]} dutch reform
\item \verb{[60044]} disciples of christ
\item \verb{[60045]} evangelical, evangelist
\item \verb{[60046]} evangelical reformed
\item \verb{[60047]} evangelist free church
\item \verb{[60048]} first church
\item \verb{[60049]} 1st christian disc of christ
\item \verb{[60050]} 1st reformed
\item \verb{[60051]} 1st christian
\item \verb{[60052]} full gospel
\item \verb{[60056]} holiness; ch of holiness
\item \verb{[60057]} pilgrim holiness
\item \verb{[60058]} jehovah witnesses
\item \verb{[60059]} lds
\item \verb{[60060]} lds--mormon
\item \verb{[60061]} lds--reorganized
\item \verb{[60062]} lds--jesus christ; ch of jesus lds
\item \verb{[60063]} mennonite
\item \verb{[60064]} mormon
\item \verb{[60065]} nazarene
\item \verb{[60066]} pentecostal assembly of god
\item \verb{[60068]} pentecostal
\item \verb{[60070]} quaker
\item \verb{[60071]} reformed
\item \verb{[60076]} salvation army
\item \verb{[60077]} 7th day adventist
\item \verb{[60078]} sanctified, sanctification
\item \verb{[60080]} unitarian, universalist
\item \verb{[60081]} united church of christ
\item \verb{[60082]} united church, unity church
\item \verb{[60083]} wesleyan
\item \verb{[60084]} wesleyan methodist--pilgrim
\item \verb{[60092]} mission covenant
\item \verb{[60093]} missionary baptist
\item \verb{[60097]} other fundamentalist
\item \verb{[60098]} federated church
\item \verb{[60102]} charismatic christian
\item \verb{[60112]} evangeligal methodist
\item \verb{[60113]} worldwide church of god
\item \verb{[60117]} missionary
\item \verb{[60119]} united church of canada
\item \verb{[60120]} evangelical united brethren
\item \verb{[60133]} primitive baptist
\item \verb{[60138]} apostolic church
\item \verb{[60141]} new birth christian
\item \verb{[60149]} polish catholic
\item \verb{[60150]} zwinglian
\item \verb{[60156]} living church of god
\item \verb{[60181]} christian born again
\item \verb{[60184]} church of god israelite
\item \verb{[70000]} no denom. given or non-denom ch
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4672/vshow}{GSS Data Explorer page for \code{densp}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab amer baptist assoc \tab amer baptist ch in usa \tab natl baptist conv. of america \tab southern baptist conv \tab other baptist chs \tab baptist, dk which \tab african meth epis ch \tab united meth church \tab methodist, dk which \tab amer lutheran ch \tab luth ch in amer \tab luth ch--mo synod \tab evan lutheran \tab lutheran, dk which \tab presbyt ch in usa \tab united presbyt ch in usa \tab other presbyt chs \tab presbyt ch usa \tab presbyterian, dk which \tab episcopal church \tab evangeligal congregational \tab ind. bible fellowship \tab christ. and mission. alliance \tab assembly of god \tab free will baptist \tab brethren church \tab united brethren, breth. in christ \tab independent \tab christian; central christian \tab christian scientist \tab church of christ \tab churches of god -exc. christ & holiness \tab church of god in christ \tab congregationalist, 1st cong \tab dutch reform \tab disciples of christ \tab evangelical, evangelist \tab evangelist free church \tab 1st christian \tab holiness; ch of holiness \tab pilgrim holiness \tab jehovah witnesses \tab lds \tab lds--mormon \tab lds--reorganized \tab lds--jesus christ; ch of jesus lds \tab mennonite \tab mormon \tab nazarene \tab pentecostal \tab quaker \tab reformed \tab 7th day adventist \tab unitarian, universalist \tab united church of christ \tab wesleyan \tab missionary baptist \tab other fundamentalist \tab federated church \tab united church of canada \tab primitive baptist \tab no denom. given or non-denom ch \tab don't know \tab no answer \tab natl baptist conv., usa, inc \tab other meth ch \tab wisc evan luth synod \tab other luth chs \tab holiness, nazarene \tab christian reform \tab evangelical reformed \tab evangelical united brethren \tab moravian \tab adventist \tab apostolic \tab missionary \tab living church of god \tab christian born again \tab church of god israelite \tab not available in this year \tab Total \cr
   1991 \tab 800 \tab 11 \tab 2 \tab 1 \tab 42 \tab 8 \tab 45 \tab 2 \tab 32 \tab 10 \tab 6 \tab 3 \tab 4 \tab 6 \tab 11 \tab 6 \tab 9 \tab 2 \tab 2 \tab 14 \tab 13 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 11 \tab 1 \tab 8 \tab 4 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 15 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 10 \tab 1 \tab 1 \tab 3 \tab 3 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 10 \tab 2 \tab 369 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1998 \tab 2473 \tab - \tab - \tab - \tab 21 \tab 7 \tab 70 \tab 12 \tab 6 \tab 28 \tab 2 \tab - \tab 4 \tab - \tab 39 \tab - \tab - \tab - \tab - \tab 12 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 8 \tab - \tab 7 \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab 11 \tab 1 \tab 3 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab 4 \tab 75 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2008 \tab 1639 \tab 5 \tab 4 \tab 2 \tab 45 \tab 1 \tab 45 \tab 1 \tab 34 \tab 7 \tab 7 \tab 1 \tab 6 \tab 2 \tab 13 \tab 3 \tab 5 \tab - \tab - \tab 15 \tab 16 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 16 \tab 1 \tab 5 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 6 \tab 2 \tab - \tab - \tab - \tab 8 \tab 4 \tab 14 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 50 \tab 8 \tab 4 \tab 4 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2023 \cr
   Total \tab 4912 \tab 16 \tab 6 \tab 3 \tab 108 \tab 16 \tab 160 \tab 15 \tab 72 \tab 45 \tab 15 \tab 4 \tab 14 \tab 8 \tab 63 \tab 9 \tab 14 \tab 2 \tab 2 \tab 41 \tab 31 \tab 1 \tab 1 \tab 5 \tab 8 \tab 4 \tab 2 \tab 1 \tab 3 \tab 35 \tab 2 \tab 20 \tab 9 \tab 3 \tab 12 \tab 1 \tab 2 \tab 4 \tab 1 \tab 5 \tab 6 \tab 2 \tab 11 \tab 21 \tab 3 \tab 1 \tab 1 \tab 1 \tab 13 \tab 9 \tab 35 \tab 4 \tab 6 \tab 9 \tab 7 \tab 7 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 65 \tab 14 \tab 448 \tab 5 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 0 \tab 6372 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{befair}},
\code{\link{bible1}},
\code{\link{cantrust}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conbiz}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{hapunhap}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{madenkid}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relcand}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}

Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{madenkid}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numbwfrd}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{othsch}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcand}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
