% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{depndabl}
\alias{depndabl}
\title{R is a dependable person}
\source{
General Social Survey https://gss.norc.org
}
\description{
depndabl
}
\details{
Question For each of the following, please indicate how well the description applies to you by circling one number:
A. A dependable person. Is this ………
}
\section{Values}{

\itemize{
\item \verb{[1]} a very good description
\item \verb{[2]} a good description
\item \verb{[3]} a fair description
\item \verb{[4]} not a very good description
\item \verb{[5]} not a good description at all
\item \verb{[9]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3047/vshow}{GSS Data Explorer page for \code{depndabl}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab a very good description \tab a good description \tab a fair description \tab not a very good description \tab not a good description at all \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2004 \tab 432 \tab 1212 \tab 920 \tab 203 \tab 23 \tab 16 \tab 1 \tab 5 \tab - \tab 2812 \cr
   Total \tab 432 \tab 1212 \tab 920 \tab 203 \tab 23 \tab 16 \tab 1 \tab 5 \tab 0 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{athletic}},
\code{\link{geneexps}},
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{genevig1}},
\code{\link{genevig2}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{owndoing}},
\code{\link{pessimst}},
\code{\link{sadblue}},
\code{\link{satself}},
\code{\link{selfish}},
\code{\link{slfrspct}}

Other Satisfaction: 
\code{\link{afailure}},
\code{\link{athletic}},
\code{\link{jobsat}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{owndoing}},
\code{\link{pessimst}},
\code{\link{sadblue}},
\code{\link{satcar}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfam7}},
\code{\link{satfin}},
\code{\link{satfixes}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{sathome}},
\code{\link{satjob}},
\code{\link{satjob7}},
\code{\link{satself}},
\code{\link{selfish}},
\code{\link{slfrspct}},
\code{\link{wktopsat}}
}
\concept{Genes and the Environment}
\concept{Satisfaction}
\keyword{variable}
