% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{depress}
\alias{depress}
\title{Told have depression}
\source{
General Social Survey https://gss.norc.org
}
\description{
depress
}
\details{
Question Now I would like to ask you some questions about general health conditions. Has a doctor, nurse, or other health professional EVER told you that you had:
D. Depression?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2845/vshow}{GSS Data Explorer page for \code{depress}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2014 \tab 1288 \tab 193 \tab 1051 \tab 6 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 271 \tab 1143 \tab 33 \tab 1 \tab - \tab 2348 \cr
   2022 \tab 1557 \tab 505 \tab 1457 \tab 8 \tab 4 \tab 13 \tab 3544 \cr
   Total \tab 3745 \tab 969 \tab 3651 \tab 47 \tab 5 \tab 13 \tab 8430 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{knowschd}},
\code{\link{satjob1}},
\code{\link{slpprblm}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}
}
\concept{Quality of Working Life}
\keyword{variable}
