% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{devil}
\alias{devil}
\title{Belief in the devil}
\source{
General Social Survey https://gss.norc.org
}
\description{
devil
}
\details{
Question Do you believe in. . .
B. The Devil
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, definitely
\item \verb{[2]} yes, probably
\item \verb{[3]} no, probably not
\item \verb{[4]} no, definitely not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4652/vshow}{GSS Data Explorer page for \code{devil}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably not \tab no, definitely not \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 550 \tab 242 \tab 204 \tab 216 \tab 65 \tab 82 \tab 1517 \cr
   Total \tab 158 \tab 550 \tab 242 \tab 204 \tab 216 \tab 65 \tab 82 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
