% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{devtype}
\alias{devtype}
\title{Are you taking this survey on a...}
\source{
General Social Survey https://gss.norc.org
}
\description{
devtype
}
\details{
Question Are you taking this survey on a…
}
\section{Values}{

\itemize{
\item \verb{[1]} laptop, netbook, or personal computer
\item \verb{[2]} smartphone or mobile phone
\item \verb{[3]} tablet
\item \verb{[4]} something else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8230/vshow}{GSS Data Explorer page for \code{devtype}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab laptop, netbook, or personal computer \tab smartphone or mobile phone \tab tablet \tab something else \tab iap \tab Total \cr
   2024 \tab - \tab 850 \tab 736 \tab 117 \tab 2 \tab 1604 \tab 3309 \cr
   Total \tab 0 \tab 850 \tab 736 \tab 117 \tab 2 \tab 1604 \tab 3309 \cr
}
}

\concept{Paradata}
\keyword{variable}
