% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_226.R
\name{diagnosd}
\alias{diagnosd}
\title{R ever diagnosed with mh problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
diagnosd
}
\details{
Question Have you ever been diagnosed with a mental health problem?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7147/vshow}{GSS Data Explorer page for \code{diagnosd}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2018 \tab 1175 \tab 197 \tab 936 \tab 5 \tab 35 \tab - \tab - \tab 2348 \cr
   2024 \tab 49 \tab 739 \tab 2462 \tab 7 \tab 15 \tab - \tab 37 \tab 3309 \cr
   Total \tab 1224 \tab 936 \tab 3398 \tab 12 \tab 50 \tab 0 \tab 37 \tab 5657 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{othmhneg}}
}
\concept{Mental Health}
\keyword{variable}
