% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{didlessp}
\alias{didlessp}
\title{Health prevent doing desired activities past 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
didlessp
}
\details{
Question Accomplished less than you would like?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2679/vshow}{GSS Data Explorer page for \code{didlessp}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 343 \tab 1067 \tab 9 \tab 2817 \cr
   Total \tab 1398 \tab 343 \tab 1067 \tab 9 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{crelesse}},
\code{\link{didlesse}},
\code{\link{limitedp}},
\code{\link{modact}},
\code{\link{stairs}}
}
\concept{Health Status}
\keyword{variable}
