% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{diffcare}
\alias{diffcare}
\title{R had difficulty getting specialty care}
\source{
General Social Survey https://gss.norc.org
}
\description{
diffcare
}
\details{
Question Have you, or anyone in your household, ever had difficulty getting a test or getting care from a specialist when you felt you needed to?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2746/vshow}{GSS Data Explorer page for \code{diffcare}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 375 \tab 416 \tab 1961 \tab 7 \tab 6 \tab 2765 \cr
   Total \tab 375 \tab 416 \tab 1961 \tab 7 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
