% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{difrace2}
\alias{difrace2}
\title{Why observation disagrees self id}
\source{
General Social Survey https://gss.norc.org
}
\description{
difrace2
}
\details{
Question A. WHY DO YOU THINK YOUR RACIAL CLASSIFICATION OF THE RESPONDENT IN "A" DIFFERED FROM THE RESPONDENT'S SELF-CLASSIFICATION IN "C"?
2. Second reason
}
\section{Values}{

\itemize{
\item \verb{[1]} r not serious
\item \verb{[2]} hispanic chosen over white
\item \verb{[3]} hispanic chosen over black
\item \verb{[4]} physically doesn't match race
\item \verb{[5]} racially mixed
\item \verb{[6]} r rejects racial categories
\item \verb{[7]} r rejects 'black' label
\item \verb{[8]} is married to member of another race
\item \verb{[9]} white or black chosen over hispanic
\item \verb{[10]} feelings, not biology matters
\item \verb{[11]} identifies with other than 'real' race
\item \verb{[12]} other hispanic complications
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5281/vshow}{GSS Data Explorer page for \code{difrace2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab physically doesn't match race \tab racially mixed \tab feelings, not biology matters \tab not available in this year \tab Total \cr
   1996 \tab 2897 \tab 3 \tab 3 \tab 1 \tab - \tab 2904 \cr
   Total \tab 2897 \tab 3 \tab 3 \tab 1 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethadapt}},
\code{\link{ethclose}},
\code{\link{ethdiff}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethnic}},
\code{\link{ethnofit}},
\code{\link{ethnum}},
\code{\link{ethorgs}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{fucitzn}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hspgovt}},
\code{\link{meltpot}},
\code{\link{natborn}},
\code{\link{owneth}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{schleth}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{teacheth}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{whoteach}},
\code{\link{whtgovt}}

Other Race: 
\code{\link{admirblk}},
\code{\link{allowmar}},
\code{\link{asngovt}},
\code{\link{blkgains}},
\code{\link{blkgovt}},
\code{\link{blkinflu}},
\code{\link{blksimp}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{bwfriend}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{color}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace3}},
\code{\link{dsndntpay}},
\code{\link{ethchng}},
\code{\link{ethhist}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdmar}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{hseth}},
\code{\link{hspgovt}},
\code{\link{hswht}},
\code{\link{leastcom}},
\code{\link{meltpot}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{numbwfrd}},
\code{\link{oprace}},
\code{\link{owneth}},
\code{\link{perla}},
\code{\link{perla_self}},
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs16}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs7}},
\code{\link{raceacs9}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmeet}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpromo}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{ractrust}},
\code{\link{racwork}},
\code{\link{ratetone}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{trtmyrac}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wrkwayup}}
}
\concept{Ethnicity}
\concept{Race}
\keyword{variable}
