% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{difrace3}
\alias{difrace3}
\title{Why observation disagrees self id}
\source{
General Social Survey https://gss.norc.org
}
\description{
difrace3
}
\details{
Question A. WHY DO YOU THINK YOUR RACIAL CLASSIFICATION OF THE RESPONDENT IN "A" DIFFERED FROM THE RESPONDENT'S SELF-CLASSIFICATION IN "C"?
3. Third reason
}
\section{Values}{

\itemize{
\item \verb{[1]} r not serious
\item \verb{[2]} hispanic chosen over white
\item \verb{[3]} hispanic chosen over black
\item \verb{[4]} physically doesn't match race
\item \verb{[5]} racially mixed
\item \verb{[6]} r rejects racial categories
\item \verb{[7]} r rejects 'black' label
\item \verb{[8]} is married to member of another race
\item \verb{[9]} white or black chosen over hispanic
\item \verb{[10]} feelings, not biology matters
\item \verb{[11]} identifies with other than 'real' race
\item \verb{[12]} other hispanic complications
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5282/vshow}{GSS Data Explorer page for \code{difrace3}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab other hispanic complications \tab Total \cr
   1996 \tab 2903 \tab 1 \tab 2904 \cr
   Total \tab 2903 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\keyword{variable}
