% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{difstand}
\alias{difstand}
\title{Some ppl hold standard in workplace that others don?}
\source{
General Social Survey https://gss.norc.org
}
\description{
difstand
}
\details{
Question A. Some people are held to standards in my workplace that others are not.
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2998/vshow}{GSS Data Explorer page for \code{difstand}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 462 \tab 599 \tab 316 \tab 409 \tab 38 \tab 43 \tab 2812 \cr
   2012 \tab 795 \tab 360 \tab 407 \tab 164 \tab 188 \tab 26 \tab 34 \tab 1974 \cr
   Total \tab 1740 \tab 822 \tab 1006 \tab 480 \tab 597 \tab 64 \tab 77 \tab 4786 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{lackinfo}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{putdown}},
\code{\link{rptprobs}}
}
\concept{Workplace Conflict}
\keyword{variable}
