% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{dinefrds}
\alias{dinefrds}
\title{How often r went out to eat or drink with at least 3 non-family members}
\source{
General Social Survey https://gss.norc.org
}
\description{
dinefrds
}
\details{
Question How often do you go out to eat or drink with three or more friends or acquaintances who are not family members?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several times a week
\item \verb{[3]} once a week
\item \verb{[4]} two to three times a month
\item \verb{[5]} once a month
\item \verb{[6]} several times a year
\item \verb{[7]} less often
\item \verb{[8]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7261/vshow}{GSS Data Explorer page for \code{dinefrds}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab daily \tab several times a week \tab once a week \tab two to three times a month \tab once a month \tab several times a year \tab less often \tab never \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 13 \tab 57 \tab 113 \tab 178 \tab 184 \tab 196 \tab 201 \tab 228 \tab 1 \tab 2 \tab 2348 \cr
   Total \tab 1175 \tab 13 \tab 57 \tab 113 \tab 178 \tab 184 \tab 196 \tab 201 \tab 228 \tab 1 \tab 2 \tab 2348 \cr
}
}

\keyword{variable}
