% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{dipged}
\alias{dipged}
\title{Diploma, ged, or other}
\source{
General Social Survey https://gss.norc.org
}
\description{
dipged
}
\details{
Question Was that a diploma awarded by your high school at graduation at the end of 12th grade, a GED awarded after you took a test, or something else?
As of 2024, DIPGED has been recoded to only account for three categories. This has been retroactively applied to all previous years of DIPGED. The original DIPGED containing data 2012 to 2022 has been renamed DIPGED_1222.
}
\section{Values}{

\itemize{
\item \verb{[1]} high school diploma
\item \verb{[2]} ged
\item \verb{[3]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/65/vshow}{GSS Data Explorer page for \code{dipged}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab high school diploma \tab ged \tab other \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2012 \tab - \tab 1543 \tab 133 \tab 8 \tab 288 \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1990 \tab 196 \tab 15 \tab 330 \tab 6 \tab 1 \tab - \tab 2538 \cr
   2016 \tab - \tab 2237 \tab 245 \tab 41 \tab 328 \tab 15 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 1891 \tab 183 \tab 6 \tab 262 \tab 6 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 3396 \tab 306 \tab 22 \tab 269 \tab 24 \tab 8 \tab 7 \tab 4032 \cr
   2022 \tab - \tab 2858 \tab 276 \tab 32 \tab 359 \tab 13 \tab 3 \tab 3 \tab 3544 \cr
   2024 \tab - \tab 2721 \tab 231 \tab 24 \tab 309 \tab 20 \tab 3 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 16636 \tab 1570 \tab 148 \tab 2145 \tab 86 \tab 16 \tab 11 \tab 20612 \cr
}
}

\seealso{
Other Core: 
\code{\link{codipged}},
\code{\link{cosector}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}

Other Education: 
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\concept{Education}
\keyword{variable}
