% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_407.R
\name{disabld4}
\alias{disabld4}
\title{R has other physical disability}
\source{
General Social Survey https://gss.norc.org
}
\description{
disabld4
}
\details{
Question Do you have any other physical disability?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3715/vshow}{GSS Data Explorer page for \code{disabld4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 1714 \tab 296 \tab 2483 \tab 2 \tab 15 \tab 4510 \cr
   Total \tab 1714 \tab 296 \tab 2483 \tab 2 \tab 15 \tab 4510 \cr
}
}

\seealso{
Other Disability: 
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld5}},
\code{\link{disabld6}},
\code{\link{disabld7}}
}
\concept{Disability}
\keyword{variable}
