% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_407.R
\name{disabld6}
\alias{disabld6}
\title{R has difficulty in learning remembering or concentrating}
\source{
General Social Survey https://gss.norc.org
}
\description{
disabld6
}
\details{
Question Because of a physical, mental, or emotional condition lasting 3 months or longer, do you have difficulty doing any of the following… Learning, remembering or concentrating?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3717/vshow}{GSS Data Explorer page for \code{disabld6}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 1714 \tab 289 \tab 2489 \tab 3 \tab 15 \tab 4510 \cr
   Total \tab 1714 \tab 289 \tab 2489 \tab 3 \tab 15 \tab 4510 \cr
}
}

\seealso{
Other Disability: 
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld4}},
\code{\link{disabld5}},
\code{\link{disabld7}}
}
\concept{Disability}
\keyword{variable}
