% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{discaffwnv}
\alias{discaffwnv}
\title{A woman won't get a job or promotion (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
discaffwnv
}
\details{
Question What do you think the chances are these days that a woman won't get a job or promotion while an equally or less qualified man gets one instead. Is this very likely, somewhat likely, somewhat unlikely, or very unlikely these days?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} very unlikely
\item \verb{[5]} no opinion (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7607/vshow}{GSS Data Explorer page for \code{discaffwnv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab very unlikely \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2689 \tab 277 \tab 631 \tab 332 \tab 97 \tab 1 \tab 1 \tab 4 \tab 4032 \cr
   2022 \tab 2951 \tab 148 \tab 262 \tab 139 \tab 41 \tab - \tab - \tab 3 \tab 3544 \cr
   Total \tab 5640 \tab 425 \tab 893 \tab 471 \tab 138 \tab 1 \tab 1 \tab 7 \tab 7576 \cr
}
}

\keyword{variable}
