% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_328.R
\name{discipln}
\alias{discipln}
\title{Situation caused by lacking discipline in the}
\source{
General Social Survey https://gss.norc.org
}
\description{
discipln
}
\details{
Question G. In your opinion, how likely is it that (John/Mary)'s situation MIGHT be caused by lacking discipline in the home?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2927/vshow}{GSS Data Explorer page for \code{discipln}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2002 \tab 1372 \tab 219 \tab 535 \tab 317 \tab 279 \tab 33 \tab 10 \tab - \tab - \tab 2765 \cr
   2024 \tab 1660 \tab 231 \tab 704 \tab 479 \tab 168 \tab 38 \tab 5 \tab - \tab 24 \tab 3309 \cr
   Total \tab 3032 \tab 450 \tab 1239 \tab 796 \tab 447 \tab 71 \tab 15 \tab 0 \tab 24 \tab 6074 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{allergic}},
\code{\link{bettrown}},
\code{\link{geneprob}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{nextdoor}},
\code{\link{raised}},
\code{\link{violtv}}

Other Children: 
\code{\link{allergic}},
\code{\link{bettrown}},
\code{\link{geneprob}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{nextdoor}},
\code{\link{raised}},
\code{\link{violtv}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
