% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_134.R
\name{discwhy3}
\alias{discwhy3}
\title{Happened at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
discwhy3
}
\details{
Question What do you think the chances are these days that a white person won't get a job or promotion while an equally or less qualified black person gets one instead?  Is this very likely, somewhat likely, or not very likely to happen these days?
A. Do you feel this way because of something that happened to you personally, because it happened to a relative, family member or close friend, because you saw it occurring at work, or because you have heard about it from the media or other sources? (CHECK ALL THAT APPLY)
3. Saw it occurring at work
}
\section{Values}{

\itemize{
\item \verb{[1]} mentioned
\item \verb{[2]} not mentioned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1254/vshow}{GSS Data Explorer page for \code{discwhy3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab don't know \tab no answer \tab Total \cr
   1990 \tab 467 \tab 189 \tab 676 \tab 18 \tab 22 \tab 1372 \cr
   Total \tab 467 \tab 189 \tab 676 \tab 18 \tab 22 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}}

Other Multiculturalism: 
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}}
}
\concept{Intergroup Relations}
\concept{Multiculturalism}
\keyword{variable}
