% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{discwk5}
\alias{discwk5}
\title{Discrimination at work in past 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
discwk5
}
\details{
Question Over the past five years, have you been discriminated against with regard to work, for instance when applying for a job, or when being considered for a pay increase or promotion?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} did not work or did not seek work
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6325/vshow}{GSS Data Explorer page for \code{discwk5}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab did not work or did not seek work \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 269 \tab 1094 \tab 93 \tab 14 \tab 7 \tab 2867 \cr
   Total \tab 1390 \tab 269 \tab 1094 \tab 93 \tab 14 \tab 7 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
