% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{disrspct}
\alias{disrspct}
\title{R is treated with less courtesy or respect than others}
\source{
General Social Survey https://gss.norc.org
}
\description{
disrspct
}
\details{
Question In your day-to-day life how often have any of the following things happened to you? You are treated with less courtesy or respect than other people.
}
\section{Values}{

\itemize{
\item \verb{[1]} almost every day
\item \verb{[2]} at least once a week
\item \verb{[3]} a few times a month
\item \verb{[4]} a few times a year
\item \verb{[5]} less than once a year
\item \verb{[6]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7158/vshow}{GSS Data Explorer page for \code{disrspct}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab almost every day \tab at least once a week \tab a few times a month \tab a few times a year \tab less than once a year \tab never \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab 774 \tab 122 \tab 199 \tab 233 \tab 399 \tab 270 \tab 338 \tab 6 \tab 7 \tab - \tab 2348 \cr
   2021 \tab 1408 \tab 136 \tab 231 \tab 327 \tab 801 \tab 552 \tab 554 \tab 1 \tab 1 \tab 21 \tab 4032 \cr
   2022 \tab 1207 \tab 170 \tab 259 \tab 309 \tab 601 \tab 446 \tab 527 \tab 9 \tab 3 \tab 13 \tab 3544 \cr
   2024 \tab 1067 \tab 195 \tab 221 \tab 290 \tab 629 \tab 376 \tab 509 \tab 7 \tab 3 \tab 12 \tab 3309 \cr
   Total \tab 4456 \tab 623 \tab 910 \tab 1159 \tab 2430 \tab 1644 \tab 1928 \tab 23 \tab 14 \tab 46 \tab 13233 \cr
}
}

\keyword{variable}
