% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{divifkd1}
\alias{divifkd1}
\title{Should feuding parents stay together for kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
divifkd1
}
\details{
Question Do you agree or disagree…
A. When there are children in the family, parents should stay together even if they don''t get along.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4409/vshow}{GSS Data Explorer page for \code{divifkd1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 36 \tab 173 \tab 231 \tab 662 \tab 248 \tab 47 \tab 50 \tab 2992 \cr
   Total \tab 1545 \tab 36 \tab 173 \tab 231 \tab 662 \tab 248 \tab 47 \tab 50 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabfst}},
\code{\link{cohabok}},
\code{\link{divbest}},
\code{\link{divnokd1}},
\code{\link{fewrksup}},
\code{\link{hubbywk1}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{singlpar}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
