% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{divlaw}
\alias{divlaw}
\title{Divorce laws}
\source{
General Social Survey https://gss.norc.org
}
\description{
divlaw
}
\details{
Question A. Should divorce in this country be easier or more difficult to obtain than it is now?
}
\section{Values}{

\itemize{
\item \code{1} easier
\item \code{2} more difficult
\item \code{3} stay as is
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
   2004 \tab A/B \tab partial \cr
   2006 \tab A/B \tab partial \cr
   2008 \tab A/B \tab partial \cr
   2010 \tab A/B \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/627/vshow}{GSS Data Explorer page for \code{divlaw}}.
}

\seealso{
Other Core: 
\code{\link{divlawy}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{spdue}},
\code{\link{sppaid}},
\code{\link{teensex}},
\code{\link{xmarsex}}

Other Divorce: 
\code{\link{divlawy}},
\code{\link{spdue}},
\code{\link{sppaid}}

Other Split Ballots: 
\code{\link{divlawy}}
}
\concept{Core}
\concept{Divorce}
\concept{Split Ballots}
\keyword{variable}
