% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{divlawnv}
\alias{divlawnv}
\title{Divorce laws (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
divlawnv
}
\details{
Question Should divorce in this country be easier or more difficult to obtain than it is now?
}
\section{Values}{

\itemize{
\item \code{1} easier
\item \code{2} more difficult
\item \code{3} stay as is (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7610/vshow}{GSS Data Explorer page for \code{divlawnv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab easier \tab more difficult \tab stay as is (phone mode only: volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2689 \tab 824 \tab 432 \tab 20 \tab 10 \tab 1 \tab 56 \tab 4032 \cr
   2022 \tab 2959 \tab 403 \tab 165 \tab - \tab - \tab - \tab 17 \tab 3544 \cr
   Total \tab 5648 \tab 1227 \tab 597 \tab 20 \tab 10 \tab 1 \tab 73 \tab 7576 \cr
}
}

\keyword{variable}
