% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{divlawv}
\alias{divlawv}
\title{Divorce laws (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
divlawv
}
\details{
Question Should divorce in this country be easier or more difficult to obtain than it is now?
}
\section{Values}{

\itemize{
\item \verb{[1]} easier
\item \verb{[2]} more difficult
\item \verb{[3]} stay as is (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7590/vshow}{GSS Data Explorer page for \code{divlawv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab easier \tab more difficult \tab stay as is (phone mode only: volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2658 \tab 460 \tab 241 \tab 657 \tab 6 \tab 1 \tab 9 \tab 4032 \cr
   2022 \tab 2998 \tab 205 \tab 66 \tab 269 \tab - \tab - \tab 6 \tab 3544 \cr
   Total \tab 5656 \tab 665 \tab 307 \tab 926 \tab 6 \tab 1 \tab 15 \tab 7576 \cr
}
}

\keyword{variable}
