% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{divlawy}
\alias{divlawy}
\title{Divorce laws-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
divlawy
}
\details{
Question B. Should divorce in this country be easier to obtain, more difficult to obtain, or stay as it is now?
}
\section{Values}{

\itemize{
\item \verb{[1]} easier
\item \verb{[2]} more difficult
\item \verb{[3]} stay same
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/628/vshow}{GSS Data Explorer page for \code{divlawy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab easier \tab more difficult \tab stay same \tab don't know \tab no answer \tab Total \cr
   1978 \tab 761 \tab 175 \tab 252 \tab 310 \tab 33 \tab 1 \tab 1532 \cr
   Total \tab 761 \tab 175 \tab 252 \tab 310 \tab 33 \tab 1 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{spdue}},
\code{\link{sppaid}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\keyword{variable}
