% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{divnokid}
\alias{divnokid}
\title{How easy should divorce be for kidless}
\source{
General Social Survey https://gss.norc.org
}
\description{
divnokid
}
\details{
Question And in general, how easy or difficult do you think the law should make it for couples without young children to get a divorce?
}
\section{Values}{

\itemize{
\item \verb{[1]} very easy
\item \verb{[2]} fairly easy
\item \verb{[3]} neither easy nor difficult
\item \verb{[4]} fairly difficult
\item \verb{[5]} very difficult
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4387/vshow}{GSS Data Explorer page for \code{divnokid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very easy \tab fairly easy \tab neither easy nor difficult \tab fairly difficult \tab very difficult \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 234 \tab 289 \tab 420 \tab 237 \tab 125 \tab 79 \tab 30 \tab 1481 \cr
   Total \tab 67 \tab 234 \tab 289 \tab 420 \tab 237 \tab 125 \tab 79 \tab 30 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
