% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{divnow}
\alias{divnow}
\title{How easy is it to get a divorce today}
\source{
General Social Survey https://gss.norc.org
}
\description{
divnow
}
\details{
Question In general, would you say that the law now makes it easy or difficult for people who want to get divorced?
}
\section{Values}{

\itemize{
\item \verb{[1]} very easy
\item \verb{[2]} fairly easy
\item \verb{[3]} neither easy nor difficult
\item \verb{[4]} fairly difficult
\item \verb{[5]} very difficult
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4386/vshow}{GSS Data Explorer page for \code{divnow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very easy \tab fairly easy \tab neither easy nor difficult \tab fairly difficult \tab very difficult \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 572 \tab 443 \tab 197 \tab 75 \tab 52 \tab 45 \tab 30 \tab 1481 \cr
   Total \tab 67 \tab 572 \tab 443 \tab 197 \tab 75 \tab 52 \tab 45 \tab 30 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
