% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{divorce5}
\alias{divorce5}
\title{Divorced in last 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
divorce5
}
\details{
Question Divorces during the last five years.
}
\section{Values}{

\itemize{
\item \code{0} none
\item \code{1} prior 4 years
\item \code{2} last year
\item \code{3} both
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/737/vshow}{GSS Data Explorer page for \code{divorce5}}.
}

\seealso{
Other Core: 
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{unemp5}}

Other Marital Status: 
\code{\link{divrel1}},
\code{\link{divrel4}}

Other Trauma: 
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{unemp5}}
}
\concept{Core}
\concept{Marital Status}
\concept{Trauma}
\keyword{variable}
