% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{divrel1}
\alias{divrel1}
\title{Relatives divorced in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
divrel1
}
\details{
Question A. Number of relatives divorced last year.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/738/vshow}{GSS Data Explorer page for \code{divrel1}}.
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{unemp5}}

Other Marital Status: 
\code{\link{divorce5}},
\code{\link{divrel4}}

Other Relatives: 
\code{\link{divrel4}}

Other Trauma: 
\code{\link{divorce5}},
\code{\link{divrel4}},
\code{\link{unemp5}}
}
\concept{Core}
\concept{Marital Status}
\concept{Relatives}
\concept{Trauma}
\keyword{variable}
