% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{dk}
\alias{dk}
\title{Don't know response to pay items}
\source{
General Social Survey https://gss.norc.org
}
\description{
dk
}
\details{
Question Think of two people doing the same kind of work. What do you personally think should be important in deciding how much to pay them?
Looking at the things below, please write ''1'' next to the thing that you think should be most important. Then write ''2'' next to the thing that you think should be next most important, and ''3'' next to the thing that you think should be third most important. Leave the other boxes blank.
I. Respondents who couldn''t choose
}
\section{Values}{

\itemize{
\item \code{1} dk chosen
\item \code{2} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4504/vshow}{GSS Data Explorer page for \code{dk}}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab dk chosen \tab not chosen \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 53 \tab 1400 \tab 84 \tab - \tab 1537 \cr
   Total \tab 0 \tab 53 \tab 1400 \tab 84 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{bosswrks}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{hiinc}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}},
\code{\link{strngun}}
}
\concept{Work Orientation}
\keyword{variable}
