% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_246.R
\name{doc15}
\alias{doc15}
\title{Doctor not really care me as a person}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc15
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
O. I doubt that my doctor really cares about me as a person.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strong disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2214/vshow}{GSS Data Explorer page for \code{doc15}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strong disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 25 \tab 199 \tab 208 \tab 758 \tab 164 \tab 24 \tab 9 \tab 2832 \cr
   2006 \tab 3073 \tab 67 \tab 175 \tab 263 \tab 586 \tab 308 \tab 25 \tab 13 \tab 4510 \cr
   Total \tab 4518 \tab 92 \tab 374 \tab 471 \tab 1344 \tab 472 \tab 49 \tab 22 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc20}},
\code{\link{prozac}}
}
\concept{Mental Health}
\keyword{variable}
