% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_333.R
\name{doc16a}
\alias{doc16a}
\title{Doctors judgment trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc16a
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
B. I trust my doctor's judgments about my medical care.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree somewhat
\item \code{3} neither agree nor disagree
\item \code{4} disgree somewhat
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2980/vshow}{GSS Data Explorer page for \code{doc16a}}.
}

\seealso{
Other Mental Health II: 
\code{\link{doc15a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{medaddct}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{solveown}}
}
\concept{Mental Health II}
\keyword{variable}
