% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_246.R
\name{doc17}
\alias{doc17}
\title{Doctor does not do everything should for me}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc17
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
Q. I feel my doctor does not do everything s/he should for my medical care.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} uncertain
\item \code{4} disagree
\item \code{5} strong disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2216/vshow}{GSS Data Explorer page for \code{doc17}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab uncertain \tab disagree \tab strong disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 26 \tab 246 \tab 204 \tab 747 \tab 130 \tab 24 \tab 10 \tab 2832 \cr
   Total \tab 1445 \tab 26 \tab 246 \tab 204 \tab 747 \tab 130 \tab 24 \tab 10 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc20}},
\code{\link{prozac}}
}
\concept{Mental Health}
\keyword{variable}
