% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_333.R
\name{doc18a}
\alias{doc18a}
\title{Doctor put medical needs above all other}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc18a
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
C. I trust my doctor to put my medical needs above all other considerations when treating my medical problems.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree somewhat
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disgree somewhat
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2981/vshow}{GSS Data Explorer page for \code{doc18a}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree somewhat \tab neither agree nor disagree \tab disgree somewhat \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 597 \tab 571 \tab 76 \tab 95 \tab 25 \tab 21 \tab 8 \tab 2765 \cr
   Total \tab 1372 \tab 597 \tab 571 \tab 76 \tab 95 \tab 25 \tab 21 \tab 8 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{medaddct}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{solveown}}
}
\concept{Mental Health II}
\keyword{variable}
