% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_246.R
\name{doc19}
\alias{doc19}
\title{Doctor a real expert in taking care problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc19
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
S. My doctor is a real expert in taking care of medical problems like mine.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strong disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2218/vshow}{GSS Data Explorer page for \code{doc19}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strong disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 118 \tab 689 \tab 357 \tab 138 \tab 15 \tab 60 \tab 10 \tab 2832 \cr
   2006 \tab 3073 \tab 287 \tab 732 \tab 299 \tab 63 \tab 15 \tab 28 \tab 13 \tab 4510 \cr
   Total \tab 4518 \tab 405 \tab 1421 \tab 656 \tab 201 \tab 30 \tab 88 \tab 23 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc20}},
\code{\link{prozac}}
}
\concept{Mental Health}
\keyword{variable}
