% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_245.R
\name{doc5}
\alias{doc5}
\title{Doctors treat patients with respect}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc5
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
E. Doctors always treat their patients with respect.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} uncertain
\item \verb{[4]} disagree
\item \verb{[5]} strong disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2204/vshow}{GSS Data Explorer page for \code{doc5}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab uncertain \tab disagree \tab strong disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 83 \tab 618 \tab 193 \tab 420 \tab 52 \tab 14 \tab 7 \tab 2832 \cr
   Total \tab 1445 \tab 83 \tab 618 \tab 193 \tab 420 \tab 52 \tab 14 \tab 7 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc2}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}}
}
\concept{Mental Health}
\keyword{variable}
