% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_245.R
\name{doc7}
\alias{doc7}
\title{Doctors avoid unnecessary patient expenses}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc7
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
G. Doctors always avoid unnecessary patient expenses.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} uncertain
\item \code{4} disagree
\item \code{5} strong disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2206/vshow}{GSS Data Explorer page for \code{doc7}}.
}

\seealso{
Other Mental Health: 
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc2}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc8}},
\code{\link{doc9}}
}
\concept{Mental Health}
\keyword{variable}
