% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{doccosts}
\alias{doccosts}
\title{How r trust doc put health above costs}
\source{
General Social Survey https://gss.norc.org
}
\description{
doccosts
}
\details{
Question How much would you trust the doctor to put your health above costs? Would you say...
}
\section{Values}{

\itemize{
\item \verb{[1]} completely
\item \verb{[2]} mostly
\item \verb{[3]} somewhat
\item \verb{[4]} a little
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2756/vshow}{GSS Data Explorer page for \code{doccosts}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab completely \tab mostly \tab somewhat \tab a little \tab not at all \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2002 \tab - \tab 362 \tab 647 \tab 748 \tab 377 \tab 585 \tab 41 \tab 5 \tab - \tab 2765 \cr
   Total \tab 0 \tab 362 \tab 647 \tab 748 \tab 377 \tab 585 \tab 41 \tab 5 \tab 0 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docrely}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{docvig}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrchng}},
\code{\link{insrlmts}},
\code{\link{regdoc}},
\code{\link{regdocyr}},
\code{\link{secopin}}

Other Health: 
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{anscitst}},
\code{\link{anydoc}},
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{crelesse}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{diffcare}},
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld4}},
\code{\link{disabld5}},
\code{\link{disabld6}},
\code{\link{disabld7}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc15a}},
\code{\link{doc16}},
\code{\link{doc16a}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc18a}},
\code{\link{doc19}},
\code{\link{doc19a}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc20a}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docinfo}},
\code{\link{doclist}},
\code{\link{docnomri}},
\code{\link{docrely}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{docvig}},
\code{\link{downblue}},
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downfeel}},
\code{\link{downqol}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotfeel}},
\code{\link{emotqol}},
\code{\link{emotrely}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{energy}},
\code{\link{findnwjb}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{health}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{helpsick}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hlthall}},
\code{\link{hlthcare}},
\code{\link{hlthcovr}},
\code{\link{hlthdays}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthinsr}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthplan}},
\code{\link{hlthtv}},
\code{\link{hlthtype}},
\code{\link{hlthwww}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hrdshp6}},
\code{\link{hyperten}},
\code{\link{insrchng}},
\code{\link{insrlmts}},
\code{\link{intmed}},
\code{\link{limitedp}},
\code{\link{misswork}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{modact}},
\code{\link{mykids}},
\code{\link{nodocidu}},
\code{\link{othkids}},
\code{\link{othplan}},
\code{\link{pain}},
\code{\link{painarms}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}},
\code{\link{peaceful}},
\code{\link{physhlth}},
\code{\link{regdoc}},
\code{\link{regdocyr}},
\code{\link{sathealt}},
\code{\link{secopin}},
\code{\link{sexpaid}},
\code{\link{slpprblm}},
\code{\link{socacts}},
\code{\link{sppart}},
\code{\link{stairs}},
\code{\link{swithlth}},
\code{\link{treat1}},
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{usedmntl}},
\code{\link{usedup}},
\code{\link{wait1}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}},
\code{\link{weight}}
}
\concept{Doctors & Patients}
\concept{Health}
\keyword{variable}
