% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{docinfo}
\alias{docinfo}
\title{Doctor provided enough info}
\source{
General Social Survey https://gss.norc.org
}
\description{
docinfo
}
\details{
Question Do you think the doctor provided enough information about financial incentives?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} maybe
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2753/vshow}{GSS Data Explorer page for \code{docinfo}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab maybe \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 711 \tab 1853 \tab 161 \tab 35 \tab 5 \tab 2765 \cr
   Total \tab 0 \tab 711 \tab 1853 \tab 161 \tab 35 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{doccosts}},
\code{\link{docnomri}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{insrchng}},
\code{\link{secopin}}
}
\concept{Doctors & Patients}
\keyword{variable}
