% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{docrely}
\alias{docrely}
\title{R relies on doctors knowledge}
\source{
General Social Survey https://gss.norc.org
}
\description{
docrely
}
\details{
Question I will read you some statements of beliefs people have.  Please look at the card and decide which answer best applies to you.
C. I prefer to rely on my doctor's knowledge and not try to find out about my condition on my own.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} moderately agree
\item \verb{[3]} slightly agree
\item \verb{[4]} slightly disagree
\item \verb{[5]} moderately disagree
\item \verb{[6]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2742/vshow}{GSS Data Explorer page for \code{docrely}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab moderately agree \tab slightly agree \tab slightly disagree \tab moderately disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 421 \tab 453 \tab 345 \tab 399 \tab 413 \tab 715 \tab 14 \tab 5 \tab 2765 \cr
   Total \tab 0 \tab 421 \tab 453 \tab 345 \tab 399 \tab 413 \tab 715 \tab 14 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docvig}},
\code{\link{regdoc}},
\code{\link{regdocyr}}
}
\concept{Doctors & Patients}
\keyword{variable}
