% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{docsat}
\alias{docsat}
\title{How satisfied r with office visit}
\source{
General Social Survey https://gss.norc.org
}
\description{
docsat
}
\details{
Question Based on what the doctor has just told you, how satisfied would you be with the office visit?  Would you rate the visit as...
}
\section{Values}{

\itemize{
\item \verb{[1]} excellent
\item \verb{[2]} very good
\item \verb{[3]} good
\item \verb{[4]} fair
\item \verb{[5]} poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2755/vshow}{GSS Data Explorer page for \code{docsat}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 89 \tab 372 \tab 759 \tab 697 \tab 814 \tab 30 \tab 4 \tab 2765 \cr
   Total \tab 0 \tab 89 \tab 372 \tab 759 \tab 697 \tab 814 \tab 30 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{doccosts}},
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docswtch}},
\code{\link{insrchng}},
\code{\link{secopin}}
}
\concept{Doctors & Patients}
\keyword{variable}
