% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{docswtch}
\alias{docswtch}
\title{Would r switch physicians}
\source{
General Social Survey https://gss.norc.org
}
\description{
docswtch
}
\details{
Question Based on what the doctor has just told you, if you could, would you switch physicians?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} maybe
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2757/vshow}{GSS Data Explorer page for \code{docswtch}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{doccosts}},
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docsat}},
\code{\link{insrchng}},
\code{\link{secopin}}

Other Health: 
\code{\link{doccosts}},
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docsat}},
\code{\link{insrchng}},
\code{\link{secopin}}
}
\concept{Doctors & Patients}
\concept{Health}
\keyword{variable}
