% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{docvig}
\alias{docvig}
\title{Vignette about the doctor}
\source{
General Social Survey https://gss.norc.org
}
\description{
docvig
}
\details{
Question Vignette number given to respondent.
}
\section{Values}{

\itemize{
\item \code{1} our physcian s group...(c4)
\item \code{2} it use to be that physicans...(c5)
\item \code{3} i can understand...(c6)
\item \code{4} i understand that...(c7)
\item \code{5} i have had a lot...(c8)
\item \code{6} i usually don t order...(c9)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2737/vshow}{GSS Data Explorer page for \code{docvig}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{regdoc}},
\code{\link{regdocyr}}
}
\concept{Doctors & Patients}
\keyword{variable}
