% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{docvst}
\alias{docvst}
\title{How often visit doctor}
\source{
General Social Survey https://gss.norc.org
}
\description{
docvst
}
\details{
Question During the past 12 months, how often did you visit or were visited by... a doctor?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} seldom
\item \code{3} sometimes
\item \code{4} often
\item \code{5} very often
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7825/vshow}{GSS Data Explorer page for \code{docvst}}.
}

\keyword{variable}
