% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{doingdys}
\alias{doingdys}
\title{How long spent on learning by doing}
\source{
General Social Survey https://gss.norc.org
}
\description{
doingdys
}
\details{
Question C. How long have you spent on learning by doing?
}
\section{Values}{

\itemize{
\item \verb{[1]} a month or less
\item \verb{[2]} month to three
\item \verb{[3]} 3 months to 6
\item \verb{[4]} 6 months to year
\item \verb{[5]} over a year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2156/vshow}{GSS Data Explorer page for \code{doingdys}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a month or less \tab month to three \tab 3 months to 6 \tab 6 months to year \tab over a year \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2098 \tab 270 \tab 86 \tab 81 \tab 59 \tab 186 \tab 1 \tab 51 \tab 2832 \cr
   Total \tab 2098 \tab 270 \tab 86 \tab 81 \tab 59 \tab 186 \tab 1 \tab 51 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
