% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{doingfin}
\alias{doingfin}
\title{Have completed this on-the-job training}
\source{
General Social Survey https://gss.norc.org
}
\description{
doingfin
}
\details{
Question B. Have you completed this informal on-the-job training?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} not yet/eventual
\item \code{3} never ends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2155/vshow}{GSS Data Explorer page for \code{doingfin}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab not yet/eventual \tab never ends \tab no answer \tab Total \cr
   1998 \tab 1937 \tab 664 \tab 29 \tab 161 \tab 41 \tab 2832 \cr
   Total \tab 1937 \tab 664 \tab 29 \tab 161 \tab 41 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
