% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_421.R
\name{done911f}
\alias{done911f}
\title{Has r reduced air travel}
\source{
General Social Survey https://gss.norc.org
}
\description{
done911f
}
\details{
Question (Do you know anyone who has done any of the following things because of terrorism since September 11th, 2001?)
Have you, or anyone you know reduced airplane travel?
FI PROBE: IF RESPONDENT ANSWERS 'YES,' ASK; 'Was this you, someone you know, or both yourself and someone you know?
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes, respondent
\item \verb{[2]} yes, someone respondent knows
\item \verb{[3]} yes, both respondent and someone respondent knows
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3853/vshow}{GSS Data Explorer page for \code{done911f}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab yes, respondent \tab yes, someone respondent knows \tab yes, both respondent and someone respondent knows \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 1098 \tab 83 \tab 103 \tab 55 \tab 8 \tab 5 \tab 2023 \cr
   Total \tab 671 \tab 1098 \tab 83 \tab 103 \tab 55 \tab 8 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911d}},
\code{\link{done911e}},
\code{\link{done911g}},
\code{\link{done911h}},
\code{\link{eff911c}},
\code{\link{eff911d}},
\code{\link{eff911e}},
\code{\link{eff911f}},
\code{\link{eff911g}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
