% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_422.R
\name{done911j}
\alias{done911j}
\title{Has r become vigilant of what is going on around r}
\source{
General Social Survey https://gss.norc.org
}
\description{
done911j
}
\details{
Question (Do you know anyone who has done any of the following things because of terrorism since September 11th, 2001?)
Have you, or anyone you know become more vigilant or aware of what is going on around them?
FI PROBE: IF RESPONDENT ANSWERS 'YES,' ASK; 'Was this you, someone you know, or both yourself and someone you know?
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes, respondent
\item \verb{[2]} yes, someone respondent knows
\item \verb{[3]} yes, both respondent and someone respondent knows
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3861/vshow}{GSS Data Explorer page for \code{done911j}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab yes, respondent \tab yes, someone respondent knows \tab yes, both respondent and someone respondent knows \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 496 \tab 358 \tab 65 \tab 404 \tab 23 \tab 6 \tab 2023 \cr
   Total \tab 671 \tab 496 \tab 358 \tab 65 \tab 404 \tab 23 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911i}},
\code{\link{done911k}},
\code{\link{done911l}},
\code{\link{done911m}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
