% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_422.R
\name{done911l}
\alias{done911l}
\title{Has r avoided tall buildings}
\source{
General Social Survey https://gss.norc.org
}
\description{
done911l
}
\details{
Question (Do you know anyone who has done any of the following things because of terrorism since September 11th, 2001?)
Have you, or anyone you know avoided tall buildings?
FI PROBE: IF RESPONDENT ANSWERS ‘YES,’ ASK; ‘Was this you, someone you know, or both yourself and someone you know?
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes, respondent
\item \code{2} yes, someone respondent knows
\item \code{3} yes, both respondent and someone respondent knows
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3865/vshow}{GSS Data Explorer page for \code{done911l}}.
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911m}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}}

Other Terrorism: 
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911m}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}}
}
\concept{Terrorism}
\concept{Terrorism Preparedness}
\keyword{variable}
