% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{doneelse}
\alias{doneelse}
\title{Specify what it is r has done}
\source{
General Social Survey https://gss.norc.org
}
\description{
doneelse
}
\details{
Question Please specify what it is you or someone you know has done.
}
\section{Values}{

\itemize{
\item \verb{[1]} research and collect more information
\item \verb{[2]} join military
\item \verb{[3]} avoid possible targets of terror
\item \verb{[4]} buy new insurance policy
\item \verb{[5]} psycho therapy
\item \verb{[6]} pray
\item \verb{[7]} more vigilant to surrounding
\item \verb{[8]} join secret service
\item \verb{[9]} buy weapons
\item \verb{[10]} vigilance and new assignments because they are already in military or secrete service
\item \verb{[11]} terror-related traiing at work
\item \verb{[12]} terror-related traiing in community
\item \verb{[13]} build safe place
\item \verb{[14]} deal with government
\item \verb{[15]} talk about it with other people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3871/vshow}{GSS Data Explorer page for \code{doneelse}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab research and collect more information \tab join military \tab avoid possible targets of terror \tab buy new insurance policy \tab psycho therapy \tab pray \tab more vigilant to surrounding \tab join secret service \tab buy weapons \tab vigilance and new assignments because they are already in military or secrete service \tab terror-related traiing at work \tab terror-related traiing in community \tab build safe place \tab deal with government \tab talk about it with other people \tab no answer \tab Total \cr
   2008 \tab 1956 \tab 10 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 9 \tab 8 \tab 3 \tab 7 \tab 2 \tab 7 \tab 3 \tab 3 \tab 2023 \cr
   Total \tab 1956 \tab 10 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 9 \tab 8 \tab 3 \tab 7 \tab 2 \tab 7 \tab 3 \tab 3 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911n}},
\code{\link{eff911m}},
\code{\link{eff911n}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
