% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_274.R
\name{doonown}
\alias{doonown}
\title{Has r figured out problems with software on own}
\source{
General Social Survey https://gss.norc.org
}
\description{
doonown
}
\details{
Question When you want to learn how to do something new with your software and you need advice on how to do it, which of the following do you do? (Name as many as are correct.) (PROMPT: Anything else?)
G. Figure it out yourself?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2498/vshow}{GSS Data Explorer page for \code{doonown}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2075 \tab 570 \tab 85 \tab 87 \tab 2817 \cr
   Total \tab 2075 \tab 570 \tab 85 \tab 87 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{askother}},
\code{\link{askwork}},
\code{\link{callcomp}},
\code{\link{manual}},
\code{\link{online}},
\code{\link{payother}},
\code{\link{wwwfrnd}},
\code{\link{wwwnei}},
\code{\link{wwwvol}}
}
\concept{Information Society}
\keyword{variable}
