% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_290.R
\name{doself}
\alias{doself}
\title{R figured out problems with web on own}
\source{
General Social Survey https://gss.norc.org
}
\description{
doself
}
\details{
Question When you want advice or assistance with surfing on the Web, finding a Web site, downloading software or some other online service, or getting your Internet connection to work, which of the following do you do? (Name as many as are correct.)
G. Figure it out yourself
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2629/vshow}{GSS Data Explorer page for \code{doself}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2043 \tab 596 \tab 76 \tab 3 \tab 47 \tab 2765 \cr
   Total \tab 2043 \tab 596 \tab 76 \tab 3 \tab 47 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{askelse}},
\code{\link{askwksch}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{cussup}},
\code{\link{manbook}},
\code{\link{payelse}},
\code{\link{processr}},
\code{\link{tecsup}}
}
\concept{Information Society}
\keyword{variable}
