% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_150.R
\name{dosports}
\alias{dosports}
\title{Played sports in lst yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
dosports
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
K. Participate in any sports activity such as softball, basketball, swimming, golf, bowling, skiing, or tennis.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1415/vshow}{GSS Data Explorer page for \code{dosports}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 901 \tab 693 \tab 1 \tab 11 \tab 1606 \cr
   Total \tab 0 \tab 901 \tab 693 \tab 1 \tab 11 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{camping}},
\code{\link{dance}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{huntfish}},
\code{\link{perform}},
\code{\link{plymusic}},
\code{\link{seemovie}},
\code{\link{usevcr}}
}
\concept{Culture}
\keyword{variable}
