% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_18.R
\name{dotthng}
\alias{dotthng}
\title{D.o.t. code - relation to things}
\source{
General Social Survey https://gss.norc.org
}
\description{
dotthng
}
\details{
Question C. Occupational Relationship with Things.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{dotged}},
\code{\link{dotpeop}},
\code{\link{dotpres}},
\code{\link{dotsvp}},
\code{\link{spdotdat}},
\code{\link{spdotged}},
\code{\link{spdotpeo}},
\code{\link{spdotsvp}},
\code{\link{spdotthn}}
}
\concept{Core}
\keyword{variable}
