% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{doubts1}
\alias{doubts1}
\title{Has evil in the world caused r doubt}
\source{
General Social Survey https://gss.norc.org
}
\description{
doubts1
}
\details{
Question How often have these problems caused doubts about your religious faith- often, sometimes, or never.
A. Evil in the world
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1095/vshow}{GSS Data Explorer page for \code{doubts1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 158 \tab 536 \tab 738 \tab 31 \tab 18 \tab 1481 \cr
   1998 \tab 1387 \tab 171 \tab 476 \tab 750 \tab 37 \tab 11 \tab 2832 \cr
   Total \tab 1387 \tab 329 \tab 1012 \tab 1488 \tab 68 \tab 29 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
