% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{doubts3}
\alias{doubts3}
\title{Has science-faith conflict caused doubt}
\source{
General Social Survey https://gss.norc.org
}
\description{
doubts3
}
\details{
Question How often have these problems caused doubts about your religious faith- often, sometimes, or never.
C. Conflict of faith and science
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1097/vshow}{GSS Data Explorer page for \code{doubts3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 73 \tab 369 \tab 955 \tab 64 \tab 20 \tab 1481 \cr
   Total \tab 0 \tab 73 \tab 369 \tab 955 \tab 64 \tab 20 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
