% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_455.R
\name{down1}
\alias{down1}
\title{Help for depression #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
down1
}
\details{
Question Now suppose you feel just a bit down or depressed, and you wanted to talk about it.
A. Who would you turn to first for help?
B. Who would you turn to second?
PLEASE CIRCLE ONLY ONE FIRST CHOICE NUMBER AND ONE SECOND CHOICE NUMBER.
A. First Choice
}
\section{Values}{

\itemize{
\item \code{1} spouse
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other relative
\item \code{9} closest friend
\item \code{10} other friend
\item \code{11} neighbor
\item \code{12} coworker
\item \code{13} clergy
\item \code{14} family doctor
\item \code{15} prof counselor
\item \code{16} other
\item \code{17} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4177/vshow}{GSS Data Explorer page for \code{down1}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab neighbor \tab coworker \tab clergy \tab family doctor \tab prof counselor \tab other \tab no one \tab no answer \tab Total \cr
   1986 \tab 42 \tab 570 \tab 75 \tab 13 \tab 73 \tab 20 \tab 87 \tab 25 \tab 24 \tab 363 \tab 17 \tab 12 \tab 9 \tab 66 \tab 14 \tab 11 \tab 4 \tab 34 \tab 11 \tab 1470 \cr
   Total \tab 42 \tab 570 \tab 75 \tab 13 \tab 73 \tab 20 \tab 87 \tab 25 \tab 24 \tab 363 \tab 17 \tab 12 \tab 9 \tab 66 \tab 14 \tab 11 \tab 4 \tab 34 \tab 11 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1}},
\code{\link{borrow2}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{down2}},
\code{\link{sick1}},
\code{\link{sick2}},
\code{\link{upset1}},
\code{\link{upset2}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
