% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_459.R
\name{down1a}
\alias{down1a}
\title{Who does r turn to first for help when depressed}
\source{
General Social Survey https://gss.norc.org
}
\description{
down1a
}
\details{
Question Now suppose you felt just a bit down or depressed, and you wanted to talk about it.
A. Who would you turn to first for help?
}
\section{Values}{

\itemize{
\item \verb{[1]} husband, wife, partner
\item \verb{[2]} mother
\item \verb{[3]} father
\item \verb{[4]} daughter
\item \verb{[5]} son
\item \verb{[6]} sister
\item \verb{[7]} brother
\item \verb{[8]} other blood relative
\item \verb{[9]} other in-law relative
\item \verb{[10]} close friend
\item \verb{[11]} neighbor
\item \verb{[12]} someone you work with
\item \verb{[13]} priest or member of the clergy
\item \verb{[14]} family doctor
\item \verb{[15]} a psychologist or other professional counselor
\item \verb{[16]} a self-help group
\item \verb{[17]} someone else
\item \verb{[18]} no one
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4218/vshow}{GSS Data Explorer page for \code{down1a}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab husband, wife, partner \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other blood relative \tab other in-law relative \tab close friend \tab neighbor \tab someone you work with \tab priest or member of the clergy \tab family doctor \tab a psychologist or other professional counselor \tab someone else \tab no one \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 363 \tab 102 \tab 21 \tab 55 \tab 14 \tab 70 \tab 32 \tab 11 \tab 13 \tab 348 \tab 5 \tab 8 \tab 3 \tab 7 \tab 11 \tab 32 \tab 44 \tab 3 \tab 7 \tab 2765 \cr
   Total \tab 1616 \tab 363 \tab 102 \tab 21 \tab 55 \tab 14 \tab 70 \tab 32 \tab 11 \tab 13 \tab 348 \tab 5 \tab 8 \tab 3 \tab 7 \tab 11 \tab 32 \tab 44 \tab 3 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1a}},
\code{\link{borrow2a}},
\code{\link{down2a}},
\code{\link{grpchrty}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{helphwrk}},
\code{\link{sick1a}},
\code{\link{sick2a}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
