% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_455.R
\name{down2}
\alias{down2}
\title{Help for depression #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
down2
}
\details{
Question Now suppose you feel just a bit down or depressed, and you wanted to talk about it.
B. Second Choice
}
\section{Values}{

\itemize{
\item \verb{[1]} spouse
\item \verb{[2]} mother
\item \verb{[3]} father
\item \verb{[4]} daughter
\item \verb{[5]} son
\item \verb{[6]} sister
\item \verb{[7]} brother
\item \verb{[8]} other relative
\item \verb{[9]} closest friend
\item \verb{[10]} other friend
\item \verb{[11]} neighbor
\item \verb{[12]} coworker
\item \verb{[13]} clergy
\item \verb{[14]} family doctor
\item \verb{[15]} prof counselor
\item \verb{[16]} other
\item \verb{[17]} no one
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4178/vshow}{GSS Data Explorer page for \code{down2}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab neighbor \tab coworker \tab clergy \tab family doctor \tab prof counselor \tab other \tab no one \tab no answer \tab Total \cr
   1986 \tab 42 \tab 82 \tab 138 \tab 29 \tab 84 \tab 63 \tab 108 \tab 64 \tab 61 \tab 349 \tab 157 \tab 34 \tab 28 \tab 66 \tab 37 \tab 29 \tab 3 \tab 71 \tab 25 \tab 1470 \cr
   Total \tab 42 \tab 82 \tab 138 \tab 29 \tab 84 \tab 63 \tab 108 \tab 64 \tab 61 \tab 349 \tab 157 \tab 34 \tab 28 \tab 66 \tab 37 \tab 29 \tab 3 \tab 71 \tab 25 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1}},
\code{\link{borrow2}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{down1}},
\code{\link{sick1}},
\code{\link{sick2}},
\code{\link{upset1}},
\code{\link{upset2}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
