% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{downblue}
\alias{downblue}
\title{Felt down and blue in past 4 wks}
\source{
General Social Survey https://gss.norc.org
}
\description{
downblue
}
\details{
Question How much of the time during the past 4 weeks have you felt downhearted and blue?
}
\section{Values}{

\itemize{
\item \code{1} all of the time
\item \code{2} most of the time
\item \code{3} a good bit of the time
\item \code{4} some of the time
\item \code{5} a little bit of the time
\item \code{6} non of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2686/vshow}{GSS Data Explorer page for \code{downblue}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab all of the time \tab most of the time \tab a good bit of the time \tab some of the time \tab a little bit of the time \tab non of the time \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 27 \tab 43 \tab 63 \tab 266 \tab 537 \tab 473 \tab 1 \tab 9 \tab 2817 \cr
   Total \tab 1398 \tab 27 \tab 43 \tab 63 \tab 266 \tab 537 \tab 473 \tab 1 \tab 9 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{energy}},
\code{\link{pain}},
\code{\link{peaceful}},
\code{\link{socacts}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}
}
\concept{Health Status}
\keyword{variable}
