% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_298.R
\name{downcure}
\alias{downcure}
\title{Treatment for downheartedness cure condition}
\source{
General Social Survey https://gss.norc.org
}
\description{
downcure
}
\details{
Question As you know, when people go to doctors, they may hold different expectations of what should result from their treatment. I’m going to describe a health problem. Please tell me how much you, yourself, would expect to get each of the following:
C. Would you definitely expect, probably not expect, or definitely not expect to be cured?
}
\section{Values}{

\itemize{
\item \code{1} definitely expect
\item \code{2} probably expect
\item \code{3} probably not expect
\item \code{4} definitely not expect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2712/vshow}{GSS Data Explorer page for \code{downcure}}.
}

\seealso{
Other Health Status: 
\code{\link{crelesse}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{downblue}},
\code{\link{downfam}},
\code{\link{downfeel}},
\code{\link{downqol}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotfeel}},
\code{\link{emotqol}},
\code{\link{emotrely}},
\code{\link{energy}},
\code{\link{limitedp}},
\code{\link{modact}},
\code{\link{pain}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}},
\code{\link{peaceful}},
\code{\link{socacts}},
\code{\link{stairs}},
\code{\link{treat1}},
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{wait1}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}}
}
\concept{Health Status}
\keyword{variable}
