% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_298.R
\name{downfam}
\alias{downfam}
\title{Treatment for downheartedness improve family}
\source{
General Social Survey https://gss.norc.org
}
\description{
downfam
}
\details{
Question As you know, when people go to doctors, they may hold different expectations of what should result from their treatment. I’m going to describe a health problem. Please tell me how much you, yourself, would expect to get each of the following:
B. Would you definitely expect, probably expect, probably not expect, or definitely not expect your relationships within your family to improve?
}
\section{Values}{

\itemize{
\item \code{1} definitely expect
\item \code{2} probably expect
\item \code{3} probably not expect
\item \code{4} definitely not expect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2711/vshow}{GSS Data Explorer page for \code{downfam}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely expect \tab probably expect \tab probably not expect \tab definitely not expect \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2362 \tab 121 \tab 195 \tab 59 \tab 28 \tab 28 \tab 24 \tab 2817 \cr
   Total \tab 2362 \tab 121 \tab 195 \tab 59 \tab 28 \tab 28 \tab 24 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downcure}},
\code{\link{downqol}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}}
}
\concept{Health Status}
\keyword{variable}
