% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{download}
\alias{download}
\title{Does r know how to download files from www}
\source{
General Social Survey https://gss.norc.org
}
\description{
download
}
\details{
Question Now we’d like to ask you a few questions about terms that people use to talk about the Internet and World Wide Web.
C. Do you know how to download a file from the World Wide Web to your computer?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2513/vshow}{GSS Data Explorer page for \code{download}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 532 \tab 131 \tab 79 \tab - \tab 2817 \cr
   2002 \tab 2142 \tab 465 \tab 110 \tab 47 \tab 1 \tab 2765 \cr
   2004 \tab 992 \tab 1478 \tab 320 \tab 19 \tab 3 \tab 2812 \cr
   Total \tab 5209 \tab 2475 \tab 561 \tab 145 \tab 4 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
