% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_299.R
\name{downrely}
\alias{downrely}
\title{Treatment for downheartedness rely less on}
\source{
General Social Survey https://gss.norc.org
}
\description{
downrely
}
\details{
Question As you know, when people go to doctors, they may hold different expectations of what should result from their treatment. I’m going to describe a health problem. Please tell me how much you, yourself, would expect to get each of the following:
D. Would you definitely expect, probably expect, probably not expect, or definitely not expect to need to rely less on others?
}
\section{Values}{

\itemize{
\item \code{1} definitely expect
\item \code{2} probably expect
\item \code{3} probably not expect
\item \code{4} definitely not expect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2713/vshow}{GSS Data Explorer page for \code{downrely}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely expect \tab probably expect \tab probably not expect \tab definitely not expect \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2362 \tab 93 \tab 170 \tab 79 \tab 45 \tab 40 \tab 28 \tab 2817 \cr
   Total \tab 2362 \tab 93 \tab 170 \tab 79 \tab 45 \tab 40 \tab 28 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downfeel}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotqol}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}}
}
\concept{Health Status}
\keyword{variable}
