% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{draft}
\alias{draft}
\title{Return to the draft?}
\source{
General Social Survey https://gss.norc.org
}
\description{
draft
}
\details{
Question Do you think we should return to a military draft at this time, or should we continue to rely on volunteers?
}
\section{Values}{

\itemize{
\item \code{1} draft
\item \code{2} volunteers
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/802/vshow}{GSS Data Explorer page for \code{draft}}.
}

\seealso{
Other Core: 
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}},
\code{\link{fehlpmil}}

Other Military: 
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}},
\code{\link{fehlpmil}}

Other Split Ballots: 
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}}
}
\concept{Core}
\concept{Military}
\concept{Split Ballots}
\keyword{variable}
