% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{draftdef}
\alias{draftdef}
\title{Defense occupations exempt from draft?}
\source{
General Social Survey https://gss.norc.org
}
\description{
draftdef
}
\details{
Question If the United States returned to a military draft, it would not be necessary to draft everyone of military age. That is, certain types of people could be exempted, even though they were otherwise qualified for service.
F. People in important defense occupations?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no, not exempt
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/811/vshow}{GSS Data Explorer page for \code{draftdef}}.
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}

Other Military: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfamnw}}

Other Split Ballots: 
\code{\link{draftco}}
}
\concept{Core}
\concept{Military}
\concept{Split Ballots}
\keyword{variable}
