% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{draftem}
\alias{draftem}
\title{If emergency, return to draft?}
\source{
General Social Survey https://gss.norc.org
}
\description{
draftem
}
\details{
Question Do you think we should return to a military draft at this time, or should we continue to rely on volunteers?
B. IF VOLUNTEERS OR DON''T KNOW: If there were a national emergency, do you think we should return to a military draft or should we continue to rely on volunteers?
}
\section{Values}{

\itemize{
\item \code{1} draft
\item \code{2} volunteers
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/804/vshow}{GSS Data Explorer page for \code{draftem}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab draft \tab volunteers \tab don't know \tab no answer \tab Total \cr
   1982 \tab 719 \tab 868 \tab 200 \tab 56 \tab 17 \tab 1860 \cr
   1983 \tab 458 \tab 928 \tab 145 \tab 42 \tab 26 \tab 1599 \cr
   1984 \tab 329 \tab 889 \tab 184 \tab 26 \tab 45 \tab 1473 \cr
   Total \tab 1506 \tab 2685 \tab 529 \tab 124 \tab 88 \tab 4932 \cr
}
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}},
\code{\link{fehlpmil}}
}
\concept{Core}
\keyword{variable}
