% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{draftmar}
\alias{draftmar}
\title{Married persons exempt from draft?}
\source{
General Social Survey https://gss.norc.org
}
\description{
draftmar
}
\details{
Question If the United States returned to a military draft, it would not be necessary to draft everyone of military age. That is, certain types of people could be exempted, even though they were otherwise qualified for service.
B. Should married persons be exempted?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no, not exempt
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/807/vshow}{GSS Data Explorer page for \code{draftmar}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no, not exempt \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 789 \tab 980 \tab 86 \tab 5 \tab 1860 \cr
   Total \tab 0 \tab 789 \tab 980 \tab 86 \tab 5 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftpar}},
\code{\link{febrass}},
\code{\link{fehlpmil}}
}
\concept{Core}
\keyword{variable}
