% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{drama}
\alias{drama}
\title{Went to a live drama}
\source{
General Social Survey https://gss.norc.org
}
\description{
drama
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
C. Go to a live performance of a non-musical stage play, not including school performances.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2131/vshow}{GSS Data Explorer page for \code{drama}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 339 \tab 1095 \tab 1 \tab 10 \tab 2832 \cr
   2002 \tab - \tab 776 \tab 1962 \tab 15 \tab 12 \tab 2765 \cr
   Total \tab 1387 \tab 1115 \tab 3057 \tab 16 \tab 22 \tab 5597 \cr
}
}

\seealso{
Other Culture: 
\code{\link{popmusic}},
\code{\link{readfict}}

Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{stateart}},
\code{\link{volarts}}
}
\concept{Culture}
\concept{Religion and Culture}
\keyword{variable}
