% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{drctlink}
\alias{drctlink}
\title{How often r links from homepage}
\source{
General Social Survey https://gss.norc.org
}
\description{
drctlink
}
\details{
Question Now I'd like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit:
A. Use a direct link from your starting page (the page that you see when you go onto the web.)?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than half of the time
\item \verb{[3]} about half of the time
\item \verb{[4]} more than half of the time
\item \verb{[5]} almost always
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2388/vshow}{GSS Data Explorer page for \code{drctlink}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than half of the time \tab about half of the time \tab more than half of the time \tab almost always \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 106 \tab 156 \tab 97 \tab 82 \tab 224 \tab 77 \tab - \tab 2817 \cr
   2002 \tab 2403 \tab 64 \tab 68 \tab 48 \tab 38 \tab 93 \tab 49 \tab 2 \tab 2765 \cr
   Total \tab 4478 \tab 170 \tab 224 \tab 145 \tab 120 \tab 317 \tab 126 \tab 2 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{srcheng}},
\code{\link{typeurl}},
\code{\link{usewww}},
\code{\link{wwwhr}},
\code{\link{wwwmin}}
}
\concept{Information Society}
\keyword{variable}
