% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{drink}
\alias{drink}
\title{Ever drink alcoholic beverages?}
\source{
General Social Survey https://gss.norc.org
}
\description{
drink
}
\details{
Question Do you ever have occasion to use any alcoholic beverages such as liquor, wine, or beer, or are you a total abstainer?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/530/vshow}{GSS Data Explorer page for \code{drink}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1977 \tab - \tab 1099 \tab 426 \tab 5 \tab 1530 \cr
   1978 \tab - \tab 1107 \tab 422 \tab 3 \tab 1532 \cr
   1980 \tab - \tab 1074 \tab 391 \tab 3 \tab 1468 \cr
   1983 \tab - \tab 1180 \tab 416 \tab 3 \tab 1599 \cr
   1984 \tab - \tab 1067 \tab 399 \tab 7 \tab 1473 \cr
   1986 \tab - \tab 991 \tab 477 \tab 2 \tab 1470 \cr
   1987 \tab - \tab 1280 \tab 534 \tab 5 \tab 1819 \cr
   1988 \tab 484 \tab 687 \tab 308 \tab 2 \tab 1481 \cr
   1989 \tab 502 \tab 691 \tab 342 \tab 2 \tab 1537 \cr
   1990 \tab 473 \tab 636 \tab 263 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 700 \tab 316 \tab 1 \tab 1517 \cr
   1993 \tab 549 \tab 733 \tab 324 \tab - \tab 1606 \cr
   1994 \tab 2481 \tab 353 \tab 158 \tab - \tab 2992 \cr
   Total \tab 4989 \tab 11598 \tab 4776 \tab 33 \tab 21396 \cr
}
}

\seealso{
Other Core: 
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{drunk}},
\code{\link{partfull}},
\code{\link{smoke}},
\code{\link{socpars}},
\code{\link{socsibs}},
\code{\link{weekswrk}}
}
\concept{Core}
\keyword{variable}
