% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{drink6up}
\alias{drink6up}
\title{How often r has 6 or more drinks in a day}
\source{
General Social Survey https://gss.norc.org
}
\description{
drink6up
}
\details{
Question Now I would like to ask you some questions about your use of alcoholic beverages during the last 12 months. How often did you have 6 or more drinks of wine, beer, or liquor in a single day? (REFERENCES AS NEEDED: that would be a bottle or more of wine/more than 2 quarts of beer/a half pint or more of liquor.)
}
\section{Values}{

\itemize{
\item \verb{[1]} never in those 12 months
\item \verb{[2]} 1 to 3 times in 12 months
\item \verb{[3]} 4 to 7 times in 12 months
\item \verb{[4]} 8 to 11 times in 12 months
\item \verb{[5]} 1 to 3 times a month
\item \verb{[6]} once or twice a week
\item \verb{[7]} 3 to 4 times a week
\item \verb{[8]} 5 times a week or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3100/vshow}{GSS Data Explorer page for \code{drink6up}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab never in those 12 months \tab 1 to 3 times in 12 months \tab 4 to 7 times in 12 months \tab 8 to 11 times in 12 months \tab 1 to 3 times a month \tab once or twice a week \tab 3 to 4 times a week \tab 5 times a week or more \tab no answer \tab Total \cr
   2004 \tab 2564 \tab 163 \tab 42 \tab 8 \tab 3 \tab 19 \tab 5 \tab 4 \tab 1 \tab 3 \tab 2812 \cr
   Total \tab 2564 \tab 163 \tab 42 \tab 8 \tab 3 \tab 19 \tab 5 \tab 4 \tab 1 \tab 3 \tab 2812 \cr
}
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink8}},
\code{\link{drinkday}},
\code{\link{drinkmax}},
\code{\link{drinkyr}},
\code{\link{evdrink}}
}
\concept{Women & Alcohol}
\keyword{variable}
