% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{drink8}
\alias{drink8}
\title{How often at least 8 but less than 12 drinks}
\source{
General Social Survey https://gss.norc.org
}
\description{
drink8
}
\details{
Question B. During the last 12 months, how often did you have at least 8, but less than 12 drinks of any kind of alcoholic beverage in a single day, that is any combination of cans, bottles or glasses of beer, glasses of wine, or drinks containing liquor of any kind. Was it:
}
\section{Values}{

\itemize{
\item \verb{[1]} never in the last 12 months
\item \verb{[2]} once in the last 12 months
\item \verb{[3]} twice in the last 12 months
\item \verb{[4]} three to six times in the last 12 months
\item \verb{[5]} seven to eleven times in the last 12 months
\item \verb{[6]} one to three times a month
\item \verb{[7]} once or twice a week
\item \verb{[8]} three or four times a week
\item \verb{[9]} every day or nearly every day
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3103/vshow}{GSS Data Explorer page for \code{drink8}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab never in the last 12 months \tab once in the last 12 months \tab twice in the last 12 months \tab three to six times in the last 12 months \tab seven to eleven times in the last 12 months \tab one to three times a month \tab once or twice a week \tab three or four times a week \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2776 \tab 1 \tab 3 \tab 6 \tab 5 \tab 4 \tab 4 \tab 4 \tab 1 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 2776 \tab 1 \tab 3 \tab 6 \tab 5 \tab 4 \tab 4 \tab 4 \tab 1 \tab 1 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drinkday}},
\code{\link{drinkmax}},
\code{\link{drinkyr}},
\code{\link{evdrink}}
}
\concept{Women & Alcohol}
\keyword{variable}
