% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{drinkday}
\alias{drinkday}
\title{How many drinks r have on a day when drinking}
\source{
General Social Survey https://gss.norc.org
}
\description{
drinkday
}
\details{
Question Again, as you think back over the last 12 months, about how many drinks would you have on a typical day when you  drank?
PROBE IF NECESSARY:  Would it have been 5 or more?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3099/vshow}{GSS Data Explorer page for \code{drinkday}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 1 \tab 12 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2564 \tab 119 \tab 1 \tab 66 \tab 27 \tab 12 \tab 8 \tab 5 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2812 \cr
   Total \tab 2564 \tab 119 \tab 1 \tab 66 \tab 27 \tab 12 \tab 8 \tab 5 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2812 \cr
}
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkmax}},
\code{\link{drinkyr}},
\code{\link{evdrink}}
}
\concept{Women & Alcohol}
\keyword{variable}
