% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{drinkmax}
\alias{drinkmax}
\title{Largest number of drinks in a day last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
drinkmax
}
\details{
Question Think of all kinds of alcoholic beverages combined, that is, any combination of cans, bottles or glasses of beer; glasses of wine; or drinks containing liquor of any kind. During the last 12 months, what is the largest number of drinks you had on any single day?  Was it:
}
\section{Values}{

\itemize{
\item \verb{[0]} did not drink at all in the last 12 months
\item \verb{[1]} at least a sip, but less than one full drink
\item \verb{[2]} at least 1, but less than 3 drinks
\item \verb{[3]} at least 3, but less than 5 drinks
\item \verb{[4]} at least 5, but less than 8 drinks
\item \verb{[5]} at least 8, but less than 12 drinks
\item \verb{[6]} at least 12, but less than 20 drinks
\item \verb{[7]} 20 or more drinks in a single day
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3101/vshow}{GSS Data Explorer page for \code{drinkmax}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab did not drink at all in the last 12 months \tab at least a sip, but less than one full drink \tab at least 1, but less than 3 drinks \tab at least 3, but less than 5 drinks \tab at least 5, but less than 8 drinks \tab at least 8, but less than 12 drinks \tab at least 12, but less than 20 drinks \tab 20 or more drinks in a single day \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2448 \tab 69 \tab 26 \tab 130 \tab 69 \tab 34 \tab 11 \tab 15 \tab 2 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 2448 \tab 69 \tab 26 \tab 130 \tab 69 \tab 34 \tab 11 \tab 15 \tab 2 \tab 1 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkday}},
\code{\link{drinkyr}},
\code{\link{evdrink}}
}
\concept{Women & Alcohol}
\keyword{variable}
