% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{drinkmin}
\alias{drinkmin}
\title{How often at least a sip but less than one full drink}
\source{
General Social Survey https://gss.norc.org
}
\description{
drinkmin
}
\details{
Question F. During the last 12 months, how often did you have at least a sip, but less than one full drink of any kind of alcoholic beverage in a single day, that is, beer, wine, or a drink containing liquor of any kind. Was it:
}
\section{Values}{

\itemize{
\item \verb{[1]} never in the last 12 months
\item \verb{[2]} once in the last 12 months
\item \verb{[3]} twice in the last 12 months
\item \verb{[4]} three to six times in the last 12 months
\item \verb{[5]} seven to eleven times in the last 12 months
\item \verb{[6]} one to three times a month
\item \verb{[7]} once or twice a week
\item \verb{[8]} three or four times a week
\item \verb{[9]} every day or nearly every day
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3107/vshow}{GSS Data Explorer page for \code{drinkmin}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab never in the last 12 months \tab once in the last 12 months \tab twice in the last 12 months \tab three to six times in the last 12 months \tab seven to eleven times in the last 12 months \tab one to three times a month \tab once or twice a week \tab three or four times a week \tab every day or nearly every day \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2448 \tab 133 \tab 37 \tab 32 \tab 43 \tab 15 \tab 37 \tab 31 \tab 16 \tab 7 \tab 6 \tab 7 \tab 2812 \cr
   Total \tab 2448 \tab 133 \tab 37 \tab 32 \tab 43 \tab 15 \tab 37 \tab 31 \tab 16 \tab 7 \tab 6 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink1}},
\code{\link{drink3}},
\code{\link{drink5}}
}
\concept{Women & Alcohol}
\keyword{variable}
