% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{drivless}
\alias{drivless}
\title{Drive less for envir reasons}
\source{
General Social Survey https://gss.norc.org
}
\description{
drivless
}
\details{
Question D. And how often do you cut back on driving a car for environmental reasons?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{5} no car, dont drive
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4772/vshow}{GSS Data Explorer page for \code{drivless}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab never \tab no car, dont drive \tab no answer \tab don't know \tab Total \cr
   1993 \tab 49 \tab 32 \tab 119 \tab 393 \tab 860 \tab 113 \tab 40 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 20 \tab 75 \tab 366 \tab 808 \tab 98 \tab 19 \tab - \tab 2992 \cr
   2010 \tab 614 \tab 64 \tab 193 \tab 427 \tab 637 \tab 98 \tab 7 \tab 4 \tab 2044 \cr
   Total \tab 2269 \tab 116 \tab 387 \tab 1186 \tab 2305 \tab 309 \tab 66 \tab 4 \tab 6642 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
