% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_137.R
\name{drugtest}
\alias{drugtest}
\title{Before got the job took a drug and alcohol test}
\source{
General Social Survey https://gss.norc.org
}
\description{
drugtest
}
\details{
Question When you started to work for your present employer (IF SELF-EMPLOYED:  current organization), did you…
C. Take a drug or alcohol test?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1287/vshow}{GSS Data Explorer page for \code{drugtest}}.
}

\seealso{
Other Work Organizations: 
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{intltest}},
\code{\link{ltrsref}},
\code{\link{physical}},
\code{\link{skiltest}}

Other Employment: 
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{intltest}},
\code{\link{ltrsref}},
\code{\link{physical}},
\code{\link{skiltest}}
}
\concept{Drinking}
\concept{Drug Use And Addiction}
\concept{Employment}
\concept{Work Organizations}
\keyword{variable}
