% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{drunk}
\alias{drunk}
\title{Ever drink too much?}
\source{
General Social Survey https://gss.norc.org
}
\description{
drunk
}
\details{
Question Do you ever have occasion to use any alcoholic beverages such as liquor, wine, or beer, or are you a total abstainer?
A. Do you sometimes drink more than you think you should?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/531/vshow}{GSS Data Explorer page for \code{drunk}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1977 \tab 426 \tab 407 \tab 677 \tab 20 \tab - \tab 1530 \cr
   1978 \tab 422 \tab 388 \tab 711 \tab 8 \tab 3 \tab 1532 \cr
   1980 \tab 391 \tab 415 \tab 642 \tab 17 \tab 3 \tab 1468 \cr
   1983 \tab 416 \tab 436 \tab 740 \tab 7 \tab - \tab 1599 \cr
   1984 \tab 399 \tab 456 \tab 598 \tab 17 \tab 3 \tab 1473 \cr
   1986 \tab 477 \tab 337 \tab 647 \tab 7 \tab 2 \tab 1470 \cr
   1987 \tab 534 \tab 500 \tab 772 \tab 13 \tab - \tab 1819 \cr
   1988 \tab 792 \tab 255 \tab 424 \tab 10 \tab - \tab 1481 \cr
   1989 \tab 844 \tab 245 \tab 435 \tab 10 \tab 3 \tab 1537 \cr
   1990 \tab 736 \tab 220 \tab 404 \tab 9 \tab 3 \tab 1372 \cr
   1991 \tab 816 \tab 201 \tab 494 \tab 2 \tab 4 \tab 1517 \cr
   1993 \tab 873 \tab 240 \tab 491 \tab 1 \tab 1 \tab 1606 \cr
   1994 \tab 2639 \tab 134 \tab 217 \tab 1 \tab 1 \tab 2992 \cr
   Total \tab 9765 \tab 4234 \tab 7252 \tab 122 \tab 23 \tab 21396 \cr
}
}

\seealso{
Other Core: 
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{drink}},
\code{\link{partfull}},
\code{\link{smoke}},
\code{\link{socpars}},
\code{\link{socsibs}},
\code{\link{weekswrk}}
}
\concept{Core}
\keyword{variable}
