% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{dsndntpay}
\alias{dsndntpay}
\title{Reparations for descendants of slaves}
\source{
General Social Survey https://gss.norc.org
}
\description{
dsndntpay
}
\details{
Question Do you think descendants of people enslaved in the U.S. should be repaid in some way, such as given land or money?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} unsure
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8238/vshow}{GSS Data Explorer page for \code{dsndntpay}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab yes \tab no \tab unsure \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 635 \tab 1114 \tab 391 \tab 37 \tab 1106 \tab 8 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 635 \tab 1114 \tab 391 \tab 37 \tab 1106 \tab 8 \tab 18 \tab 3309 \cr
}
}

\concept{Race}
\keyword{variable}
