% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{dwelcity}
\alias{dwelcity}
\title{Dwelling compared to others in city}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwelcity
}
\details{
Question B. Compared to houses/apartments in the city/town/county, would you say the house was . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} far above average
\item \verb{[2]} above average
\item \verb{[3]} average
\item \verb{[4]} below average
\item \verb{[5]} far below average
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5293/vshow}{GSS Data Explorer page for \code{dwelcity}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab far above average \tab above average \tab average \tab below average \tab far below average \tab no answer \tab Total \cr
   1982 \tab - \tab 37 \tab 318 \tab 982 \tab 382 \tab 72 \tab 69 \tab 1860 \cr
   1983 \tab - \tab 43 \tab 312 \tab 872 \tab 300 \tab 44 \tab 28 \tab 1599 \cr
   1984 \tab - \tab 34 \tab 272 \tab 810 \tab 273 \tab 38 \tab 46 \tab 1473 \cr
   Total \tab 0 \tab 114 \tab 902 \tab 2664 \tab 955 \tab 154 \tab 143 \tab 4932 \cr
}
}

\seealso{
Other Residence: 
\code{\link{dwelling}},
\code{\link{dwelngh}}
}
\concept{Residence}
\keyword{variable}
