% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{dwell5}
\alias{dwell5}
\title{In last 5 yrs, r has been prevented from moving bc landlord believed r was gay/l}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwell5
}
\details{
Question In the last five years, have you been forced to move or prevented from buying or moving into a house or apartment  because a landlord or realtor believed you were gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3927/vshow}{GSS Data Explorer page for \code{dwell5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 1995 \tab 1 \tab 1 \tab 26 \tab - \tab 2023 \cr
   2012 \tab 1925 \tab 1 \tab 2 \tab 46 \tab - \tab 1974 \cr
   Total \tab 3920 \tab 2 \tab 3 \tab 72 \tab 0 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evkid}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{mykids}},
\code{\link{negjob5}},
\code{\link{othkids}},
\code{\link{relgendr}},
\code{\link{sexsex18}},
\code{\link{sexunion}},
\code{\link{sppart}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}}

Other Dwelling (R's): 
\code{\link{bornhome}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{geomobil}},
\code{\link{huclean}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}}

Other Gays And Lesbians: 
\code{\link{acqfmgay}},
\code{\link{acqgay}},
\code{\link{acqnhgay}},
\code{\link{acqvagay}},
\code{\link{acqwkgay}},
\code{\link{asksexor}},
\code{\link{attractd}},
\code{\link{colhomo}},
\code{\link{draftgay}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{libhomo}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{negjob5}},
\code{\link{sexornt}},
\code{\link{sexsex18}},
\code{\link{sexunion}},
\code{\link{spkhomo}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}},
\code{\link{trtgay}}

Other Residence: 
\code{\link{abroad}},
\code{\link{agecmeus}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{comamind}},
\code{\link{comtype}},
\code{\link{comyear}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{geomobil}},
\code{\link{hspasn10}},
\code{\link{liveasns}},
\code{\link{livecom}},
\code{\link{livecom1}},
\code{\link{livehome}},
\code{\link{livehsps}},
\code{\link{loclived}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{res16}},
\code{\link{res161}},
\code{\link{res2006}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Dwelling (R's)}
\concept{Gays And Lesbians}
\concept{Residence}
\concept{Sexual Orientation}
\keyword{variable}
