% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{dwelling}
\alias{dwelling}
\title{Type of structure}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwelling
}
\details{
Question DWELLING TYPE.
}
\section{Values}{

\itemize{
\item \verb{[1]} trailer
\item \verb{[2]} single family house
\item \verb{[3]} multifamily house
\item \verb{[4]} apartment
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5291/vshow}{GSS Data Explorer page for \code{dwelling}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab trailer \tab single family house \tab multifamily house \tab apartment \tab other \tab no answer \tab don't know \tab iap \tab Total \cr
   1982 \tab - \tab 120 \tab 1155 \tab 293 \tab 264 \tab 11 \tab 17 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 82 \tab 1065 \tab 214 \tab 211 \tab 12 \tab 15 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 86 \tab 935 \tab 198 \tab 225 \tab 17 \tab 12 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 118 \tab 935 \tab 194 \tab 253 \tab 15 \tab 19 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 95 \tab 936 \tab 189 \tab 211 \tab 19 \tab 20 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 96 \tab 1140 \tab 237 \tab 305 \tab 27 \tab 14 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 113 \tab 848 \tab 242 \tab 241 \tab 10 \tab 27 \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 117 \tab 1006 \tab 166 \tab 182 \tab 39 \tab 27 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 84 \tab 879 \tab 151 \tab 222 \tab 23 \tab 12 \tab 1 \tab - \tab 1372 \cr
   1991 \tab - \tab 95 \tab 953 \tab 211 \tab 235 \tab 8 \tab 14 \tab 1 \tab - \tab 1517 \cr
   1993 \tab - \tab 108 \tab 1043 \tab 191 \tab 250 \tab 13 \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 195 \tab 1790 \tab 414 \tab 528 \tab 19 \tab 15 \tab 31 \tab - \tab 2992 \cr
   1996 \tab - \tab 218 \tab 1700 \tab 384 \tab 530 \tab 30 \tab 30 \tab 12 \tab - \tab 2904 \cr
   1998 \tab - \tab 173 \tab 1609 \tab 371 \tab 627 \tab 26 \tab 22 \tab 4 \tab - \tab 2832 \cr
   2000 \tab - \tab 193 \tab 1574 \tab 408 \tab 561 \tab 20 \tab 57 \tab 4 \tab - \tab 2817 \cr
   2002 \tab - \tab 197 \tab 1492 \tab 337 \tab 658 \tab 52 \tab - \tab 29 \tab - \tab 2765 \cr
   2004 \tab - \tab 158 \tab 1841 \tab 329 \tab 405 \tab 23 \tab 9 \tab 47 \tab - \tab 2812 \cr
   2006 \tab - \tab 299 \tab 2853 \tab 524 \tab 726 \tab 65 \tab 7 \tab 36 \tab - \tab 4510 \cr
   2008 \tab - \tab 118 \tab 1336 \tab 210 \tab 305 \tab 23 \tab 3 \tab 28 \tab - \tab 2023 \cr
   2010 \tab - \tab 107 \tab 1244 \tab 256 \tab 349 \tab 25 \tab 8 \tab 55 \tab - \tab 2044 \cr
   2012 \tab - \tab 128 \tab 1269 \tab 236 \tab 331 \tab 4 \tab 6 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 116 \tab 1525 \tab 229 \tab 448 \tab 23 \tab 6 \tab 191 \tab - \tab 2538 \cr
   2016 \tab - \tab 150 \tab 1840 \tab 333 \tab 528 \tab 14 \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 138 \tab 1519 \tab 272 \tab 412 \tab 6 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   Total \tab 4032 \tab 3304 \tab 32487 \tab 6589 \tab 9007 \tab 524 \tab 344 \tab 439 \tab 0 \tab 56726 \cr
}
}

\seealso{
Other Residence: 
\code{\link{dwelcity}},
\code{\link{dwelngh}}

Other Dwelling (R's): 
\code{\link{dwelcity}},
\code{\link{dwelngh}}
}
\concept{Dwelling (R's)}
\concept{Residence}
\keyword{variable}
