% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{dwelngh}
\alias{dwelngh}
\title{Dwelling compared to others in nbrhd}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwelngh
}
\details{
Question A. Compared to houses/apartments in the neighborhood, would you say the house/ apartment was . . .
}
\section{Values}{

\itemize{
\item \code{1} far above average
\item \code{2} above average
\item \code{3} average
\item \code{4} below average
\item \code{5} far below average
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5292/vshow}{GSS Data Explorer page for \code{dwelngh}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab far above average \tab above average \tab average \tab below average \tab far below average \tab no answer \tab Total \cr
   1982 \tab - \tab 40 \tab 260 \tab 1247 \tab 219 \tab 34 \tab 60 \tab 1860 \cr
   1983 \tab - \tab 30 \tab 233 \tab 1128 \tab 162 \tab 24 \tab 22 \tab 1599 \cr
   1984 \tab - \tab 19 \tab 225 \tab 1039 \tab 143 \tab 15 \tab 32 \tab 1473 \cr
   Total \tab 0 \tab 89 \tab 718 \tab 3414 \tab 524 \tab 73 \tab 114 \tab 4932 \cr
}
}

\seealso{
Other Residence: 
\code{\link{dwelcity}},
\code{\link{dwelling}}
}
\concept{Residence}
\keyword{variable}
