% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{earndes}
\alias{earndes}
\title{R deserves earnings}
\source{
General Social Survey https://gss.norc.org
}
\description{
earndes
}
\details{
Question Would you say that you earn:
}
\section{Values}{

\itemize{
\item \verb{[1]} much less than i deserve
\item \verb{[2]} less than i deserve
\item \verb{[3]} what i deserve
\item \verb{[4]} more than i deserve
\item \verb{[5]} much more than i deserve
\item \verb{[6]} never worked
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4323/vshow}{GSS Data Explorer page for \code{earndes}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab much less than i deserve \tab less than i deserve \tab what i deserve \tab more than i deserve \tab much more than i deserve \tab never worked \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 188 \tab 471 \tab 407 \tab 34 \tab 4 \tab 17 \tab 68 \tab 83 \tab 2817 \cr
   Total \tab 1545 \tab 188 \tab 471 \tab 407 \tab 34 \tab 4 \tab 17 \tab 68 \tab 83 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
