% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{earthsun}
\alias{earthsun}
\title{Sci knowledge: the earth goes around the sun}
\source{
General Social Survey https://gss.norc.org
}
\description{
earthsun
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don't know or aren't sure, just tell me so,  and we will skip to the next question. Remember true, false, or don't know.
J. Now, does the Earth go around the Sun, or does the Sun go around the Earth?
}
\section{Values}{

\itemize{
\item \verb{[1]} earth around sun
\item \verb{[2]} sun around earth
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3343/vshow}{GSS Data Explorer page for \code{earthsun}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab earth around sun \tab sun around earth \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1372 \tab 342 \tab 149 \tab 1 \tab 4510 \cr
   2008 \tab 518 \tab 1073 \tab 311 \tab 112 \tab 9 \tab 2023 \cr
   2010 \tab 1103 \tab 699 \tab 169 \tab 71 \tab 2 \tab 2044 \cr
   2012 \tab 972 \tab 707 \tab 203 \tab 84 \tab 8 \tab 1974 \cr
   2014 \tab 1299 \tab 926 \tab 247 \tab 63 \tab 3 \tab 2538 \cr
   2016 \tab 1477 \tab 987 \tab 314 \tab 82 \tab 7 \tab 2867 \cr
   2018 \tab 1173 \tab 836 \tab 281 \tab 54 \tab 4 \tab 2348 \cr
   Total \tab 9188 \tab 6600 \tab 1867 \tab 615 \tab 34 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
