% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{eatout}
\alias{eatout}
\title{Last week had a meal at a restaurant}
\source{
General Social Survey https://gss.norc.org
}
\description{
eatout
}
\details{
Question Now I'm going to ask you about things you did during the last seven days. I'm only interested in what you did during the last seven days.
From last (DAY OF WEEK) to today did you…
B. Have a meal (breakfast, lunch or dinner) at a restaurant (including fast food places and take-outs)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5307/vshow}{GSS Data Explorer page for \code{eatout}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1996 \tab 1953 \tab 765 \tab 184 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 1387 \tab 1158 \tab 282 \tab 4 \tab 1 \tab 2832 \cr
   Total \tab 3340 \tab 1923 \tab 466 \tab 6 \tab 1 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attrelig}},
\code{\link{godoc}},
\code{\link{monday}},
\code{\link{numdays}},
\code{\link{seefilm}},
\code{\link{sunday}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
