% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_374.R
\name{ecobsttx}
\alias{ecobsttx}
\title{Best policy vs narrow interest on fed inc tax: economist}
\source{
General Social Survey https://gss.norc.org
}
\description{
ecobsttx
}
\details{
Question Please look at Card B14. When making policy recommendations about federal income taxes, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
A. Economists
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3384/vshow}{GSS Data Explorer page for \code{ecobsttx}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3574 \tab 211 \tab 259 \tab 231 \tab 90 \tab 84 \tab 61 \tab - \tab 4510 \cr
   2010 \tab 1578 \tab 80 \tab 133 \tab 135 \tab 49 \tab 42 \tab 23 \tab 4 \tab 2044 \cr
   Total \tab 5152 \tab 291 \tab 392 \tab 366 \tab 139 \tab 126 \tab 84 \tab 4 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbsttx}},
\code{\link{bizinftx}},
\code{\link{ecoagree}},
\code{\link{ecoinftx}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{polbsttx}},
\code{\link{polinftx}},
\code{\link{txpol}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
