% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{ecomhro}
\alias{ecomhro}
\title{Hours per week using other computer hsehld mngm}
\source{
General Social Survey https://gss.norc.org
}
\description{
ecomhro
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
A. Shopping, paying bills and other household management?
2. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2586/vshow}{GSS Data Explorer page for \code{ecomhro}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab no answer \tab Total \cr
   2000 \tab 2665 \tab 88 \tab 6 \tab 2 \tab 1 \tab 55 \tab 2817 \cr
   Total \tab 2665 \tab 88 \tab 6 \tab 2 \tab 1 \tab 55 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Information Society}
\keyword{variable}
