% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{ecomminh}
\alias{ecomminh}
\title{Minutes per week use computer household management}
\source{
General Social Survey https://gss.norc.org
}
\description{
ecomminh
}
\details{
Question About how many minutes or hours a week do you use your home computer for each of the following:
A. Shopping, paying bills and other household management?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2556/vshow}{GSS Data Explorer page for \code{ecomminh}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 15 \tab 20 \tab 30 \tab 45 \tab 5 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2418 \tab 314 \tab 8 \tab 11 \tab 5 \tab 20 \tab 1 \tab 3 \tab 3 \tab 34 \tab 2817 \cr
   Total \tab 2418 \tab 314 \tab 8 \tab 11 \tab 5 \tab 20 \tab 1 \tab 3 \tab 3 \tab 34 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
