% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{ecommino}
\alias{ecommino}
\title{Minutes per week using other computer hsehld mn}
\source{
General Social Survey https://gss.norc.org
}
\description{
ecommino
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
A. Shopping, paying bills and other household management?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2585/vshow}{GSS Data Explorer page for \code{ecommino}}.
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{ecomhro}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}

Other Computers And Internet: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{ecomhro}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
