% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{ecomminw}
\alias{ecomminw}
\title{Minutes per week using work computer hsehld mng}
\source{
General Social Survey https://gss.norc.org
}
\description{
ecomminw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
C. Shopping, paying bills and other household management?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2571/vshow}{GSS Data Explorer page for \code{ecomminw}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 15 \tab 20 \tab 30 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2486 \tab 268 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 53 \tab 2817 \cr
   Total \tab 2486 \tab 268 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 53 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{emhrw}},
\code{\link{emminw}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{perhrw}},
\code{\link{perminw}},
\code{\link{wwwhrw}},
\code{\link{wwwminw}}
}
\concept{Information Society}
\keyword{variable}
