% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{econfutr}
\alias{econfutr}
\title{Predictions for national economy in next year}
\source{
General Social Survey https://gss.norc.org
}
\description{
econfutr
}
\details{
Question What about the next 12 months? Do you expect the national economy to get better, get worse, or stay about the same?
}
\section{Values}{

\itemize{
\item \verb{[1]} get better
\item \verb{[2]} stay the same
\item \verb{[3]} get worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1644/vshow}{GSS Data Explorer page for \code{econfutr}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab get better \tab stay the same \tab get worse \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1518 \tab 416 \tab 768 \tab 241 \tab 39 \tab 10 \tab - \tab 2992 \cr
   Total \tab 1518 \tab 416 \tab 768 \tab 241 \tab 39 \tab 10 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{asngovt}},
\code{\link{bilinged}},
\code{\link{blkgovt}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{econpast}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hspasn10}},
\code{\link{hspgovt}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{jobaff}},
\code{\link{letin}},
\code{\link{meltpot}},
\code{\link{obrespct}},
\code{\link{owneth}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{united}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Economy: 
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{bizbsttx}},
\code{\link{bizinftx}},
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{econpast}},
\code{\link{econsys}},
\code{\link{equal5}},
\code{\link{equalize}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{govcare}},
\code{\link{hlphitec}},
\code{\link{hlthcare}},
\code{\link{infljobs}},
\code{\link{intecon}},
\code{\link{jobsall}},
\code{\link{knwecon}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{moretrde}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{polbsttx}},
\code{\link{polinftx}},
\code{\link{pricecon}},
\code{\link{reqinfo}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{txbiz}},
\code{\link{txeco}},
\code{\link{txpol}}
}
\concept{Economy}
\concept{Multiculturalism}
\keyword{variable}
