% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{econgrn}
\alias{econgrn}
\title{Economic progress dependent on environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
econgrn
}
\details{
Question How much do you agree or disagree with each of these statements?
C. Economic progress in America will slow down unless we look after the environment better.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4789/vshow}{GSS Data Explorer page for \code{econgrn}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{excldimm}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{ldcgrn}},
\code{\link{nukeacc}}

Other Environment: 
\code{\link{ldcgrn}},
\code{\link{nukeacc}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
