% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_175.R
\name{econpast}
\alias{econpast}
\title{Has national economy improved last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
econpast
}
\details{
Question How about the economy. Would you say that over the past year the nation's economy has gotten better, stayed about the same, or gotten worse?
}
\section{Values}{

\itemize{
\item \verb{[1]} gotten better
\item \verb{[2]} stayed the same
\item \verb{[3]} gotten worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1641/vshow}{GSS Data Explorer page for \code{econpast}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab gotten better \tab stayed the same \tab gotten worse \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 381 \tab 636 \tab 428 \tab 20 \tab 9 \tab 2992 \cr
   Total \tab 1518 \tab 381 \tab 636 \tab 428 \tab 20 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{hspgovt}},
\code{\link{obrespct}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Multiculturalism}
\keyword{variable}
