% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{econstat}
\alias{econstat}
\title{What do you think about the state of the economy these days in the united states?}
\source{
General Social Survey https://gss.norc.org
}
\description{
econstat
}
\details{
Question What do you think about the state of the economy these days in the United States? Would you say the state of the economy is very good, good, neither good nor bad, bad, or very bad?
}
\section{Values}{

\itemize{
\item \verb{[1]} very good
\item \verb{[2]} good
\item \verb{[3]} neither bad nor good
\item \verb{[4]} bad
\item \verb{[5]} very bad
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8105/vshow}{GSS Data Explorer page for \code{econstat}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab very good \tab good \tab neither bad nor good \tab bad \tab very bad \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 102 \tab 503 \tab 727 \tab 1105 \tab 811 \tab 13 \tab 16 \tab 9 \tab 23 \tab 3309 \cr
   Total \tab 0 \tab 102 \tab 503 \tab 727 \tab 1105 \tab 811 \tab 13 \tab 16 \tab 9 \tab 23 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{aneseligible}},
\code{\link{anespost_status}},
\code{\link{anespre_status}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovote24a}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
