% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{econsup}
\alias{econsup}
\title{Main source of economic support}
\source{
General Social Survey https://gss.norc.org
}
\description{
econsup
}
\details{
Question What is your main source of economic support?
}
\section{Values}{

\itemize{
\item \code{1} pnsion pri/st/ss
\item \code{2} unemploy benfits
\item \code{3} spouse/partner
\item \code{4} othr famly mmbrs
\item \code{5} soc assis/welfre
\item \code{6} occasional work
\item \code{7} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4605/vshow}{GSS Data Explorer page for \code{econsup}}.
}

\seealso{
Other Work Orientation: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{rimpskls}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}

Other Unemployment: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{wantads}}
}
\concept{Unemployment}
\concept{Work Orientation}
\keyword{variable}
