% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{econsup1}
\alias{econsup1}
\title{Main source of economic support}
\source{
General Social Survey https://gss.norc.org
}
\description{
econsup1
}
\details{
Question (HAND CARD D24)            
What is your main source of economic support?            
Please select one source.
}
\section{Values}{

\itemize{
\item \code{1} current spouse/partner
\item \code{2} other family members
\item \code{3} pension (private or government)
\item \code{4} unemployment benefits
\item \code{5} social assistance/welfare
\item \code{6} occasional work
\item \code{7} student loans/stipends
\item \code{8} savings
\item \code{9} investments or rental income
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6349/vshow}{GSS Data Explorer page for \code{econsup1}}.
}

\seealso{
Other Work Orientation: 
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{worecsup}}
}
\concept{Work Orientation}
\keyword{variable}
