% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{econsys}
\alias{econsys}
\title{Best-needs replaced}
\source{
General Social Survey https://gss.norc.org
}
\description{
econsys
}
\details{
Question On the whole, do you think our economic system is...
}
\section{Values}{

\itemize{
\item \verb{[1]} the best possible
\item \verb{[2]} ok but need some tinkering
\item \verb{[3]} need fundamental changes
\item \verb{[4]} need to be replaced
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1879/vshow}{GSS Data Explorer page for \code{econsys}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab the best possible \tab ok but need some tinkering \tab need fundamental changes \tab need to be replaced \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1460 \tab 104 \tab 639 \tab 537 \tab 113 \tab 43 \tab 8 \tab - \tab 2904 \cr
   Total \tab 1460 \tab 104 \tab 639 \tab 537 \tab 113 \tab 43 \tab 8 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{borrowed}},
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{evbuyhme}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{fixer1}},
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{fixhome}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{haggle}},
\code{\link{homeyear}},
\code{\link{hrfambiz}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal3}},
\code{\link{inequal5}},
\code{\link{lawyer}},
\code{\link{lawyer1}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{layoffs}},
\code{\link{learnhme}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{natrecon}},
\code{\link{newowned}},
\code{\link{newused}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{realtor1}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{reldealr}},
\code{\link{relfixer}},
\code{\link{relhome}},
\code{\link{rellaw}},
\code{\link{relprivt}},
\code{\link{reqinfo}},
\code{\link{satcar}},
\code{\link{satfixes}},
\code{\link{sathome}},
\code{\link{satlawyr}},
\code{\link{sellauto}},
\code{\link{sellbaby}},
\code{\link{sellbed}},
\code{\link{sellhome}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{tagsales}},
\code{\link{transoth}},
\code{\link{transrel}},
\code{\link{typdealr}},
\code{\link{warrntyc}},
\code{\link{warrntyh}},
\code{\link{wholoand}},
\code{\link{whosold}},
\code{\link{wkfambiz}}

Other Economy: 
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{bizbsttx}},
\code{\link{bizinftx}},
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{econfutr}},
\code{\link{econpast}},
\code{\link{equal5}},
\code{\link{equalize}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{govcare}},
\code{\link{hlphitec}},
\code{\link{hlthcare}},
\code{\link{infljobs}},
\code{\link{intecon}},
\code{\link{jobsall}},
\code{\link{knwecon}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{moretrde}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{polbsttx}},
\code{\link{polinftx}},
\code{\link{pricecon}},
\code{\link{reqinfo}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{txbiz}},
\code{\link{txeco}},
\code{\link{txpol}}
}
\concept{Economy}
\concept{Markets}
\keyword{variable}
