% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{edcskill}
\alias{edcskill}
\title{How important school in developing skills}
\source{
General Social Survey https://gss.norc.org
}
\description{
edcskill
}
\details{
Question Now think about the skills that you actually use in your job. How important would you say each of the following  was in developing these skills?
A. School, college, or university
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} not un/important
\item \verb{[4]} not important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4583/vshow}{GSS Data Explorer page for \code{edcskill}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab not un/importnt \tab not important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 218 \tab 271 \tab 127 \tab 101 \tab 82 \tab 9 \tab 57 \tab 2832 \cr
   Total \tab 1967 \tab 218 \tab 271 \tab 127 \tab 101 \tab 82 \tab 9 \tab 57 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
