% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{educ2}
\alias{educ2}
\title{Education level of person #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
educ2
}
\details{
Question This card (CARD BH) lists general levels of education. As far as you know, what is (NAME)''s highest level of education?
PROBE:  Your best guess. RECORD VERBATIM IF NOT CODEABLE. ASK FOR EACH NAME.
2. Education of Second Person
}
\section{Values}{

\itemize{
\item \verb{[0]} 1-6 years
\item \verb{[1]} 7-9 years
\item \verb{[2]} 10-12 years
\item \verb{[3]} h.s. grad
\item \verb{[4]} some college
\item \verb{[5]} asso. degree
\item \verb{[6]} bach. degree
\item \verb{[7]} grad or prof
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/941/vshow}{GSS Data Explorer page for \code{educ2}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 1-6 years \tab 7-9 years \tab 10-12 years \tab h.s. grad \tab some college \tab asso. degree \tab bach. degree \tab grad or prof \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1985 \tab 364 \tab 17 \tab 70 \tab 107 \tab 393 \tab 182 \tab 28 \tab 179 \tab 143 \tab 38 \tab 13 \tab - \tab 1534 \cr
   2004 \tab 1977 \tab 4 \tab 17 \tab 41 \tab 213 \tab 143 \tab 57 \tab 177 \tab 115 \tab 19 \tab 49 \tab - \tab 2812 \cr
   Total \tab 2341 \tab 21 \tab 87 \tab 148 \tab 606 \tab 325 \tab 85 \tab 356 \tab 258 \tab 57 \tab 62 \tab 0 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{acqasian}},
\code{\link{acqattnd}},
\code{\link{acqblack}},
\code{\link{acqbrnda}},
\code{\link{acqchild}},
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqdems}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhme}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmno}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqfmwht}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjans}},
\code{\link{acqjose}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqlinda}},
\code{\link{acqmaria}},
\code{\link{acqmark}},
\code{\link{acqmils}},
\code{\link{acqmyrac}},
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhlib}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}},
\code{\link{acqnhwht}},
\code{\link{acqnoatt}},
\code{\link{acqprisn}},
\code{\link{acqrachl}},
\code{\link{acqreps}},
\code{\link{acqshawn}},
\code{\link{acqsocs}},
\code{\link{acqunemp}},
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvalib}},
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhme}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkno}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}},
\code{\link{acqwkwht}},
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{comamind}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{educ1}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{fricall}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndknow}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{grpother}},
\code{\link{grpsprts}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{hlpinflu}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{lentto}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{peocntct}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkedto}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}},
\code{\link{trtasian}},
\code{\link{trtattnd}},
\code{\link{trtblack}},
\code{\link{trtbrnda}},
\code{\link{trtchild}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtcops}},
\code{\link{trtdems}},
\code{\link{trtelecs}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtjose}},
\code{\link{trtkaren}},
\code{\link{trtkeith}},
\code{\link{trtkevin}},
\code{\link{trtlaws}},
\code{\link{trtlib}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtmils}},
\code{\link{trtmyrac}},
\code{\link{trtnoatt}},
\code{\link{trtprisn}},
\code{\link{trtrachl}},
\code{\link{trtreps}},
\code{\link{trtshawn}},
\code{\link{trtsocs}},
\code{\link{trtunemp}},
\code{\link{trtwhite}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{youinflu}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}

Other Education: 
\code{\link{actgreek}},
\code{\link{actschl}},
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{bilinged}},
\code{\link{cathsch}},
\code{\link{chldeduc}},
\code{\link{chrissch}},
\code{\link{classics}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{coldeg1}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{colop}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{datesch}},
\code{\link{degree}},
\code{\link{denomsch}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{educ}},
\code{\link{educ1}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{educemp}},
\code{\link{educop}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{govedop}},
\code{\link{homesch}},
\code{\link{hsbasics}},
\code{\link{hsbio}},
\code{\link{hscaring}},
\code{\link{hschem}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsmath}},
\code{\link{hsorder}},
\code{\link{hsphys}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inteduc}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{majorcol}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memschl}},
\code{\link{natdeg}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{opeduc}},
\code{\link{oppared}},
\code{\link{othsch}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{relsch}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{schleth}},
\code{\link{schlhead}},
\code{\link{schoolng}},
\code{\link{scimath}},
\code{\link{sexeduc}},
\code{\link{solgreek}},
\code{\link{solschl}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{speduc}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{teacheth}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{trstprof}},
\code{\link{undoccol}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{whoteach}}

Other Social Networks: 
\code{\link{acqasian}},
\code{\link{acqattnd}},
\code{\link{acqblack}},
\code{\link{acqbrnda}},
\code{\link{acqchild}},
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqdems}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhme}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmno}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqfmwht}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjans}},
\code{\link{acqjose}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqlinda}},
\code{\link{acqmaria}},
\code{\link{acqmark}},
\code{\link{acqmils}},
\code{\link{acqmyrac}},
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhlib}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}},
\code{\link{acqnhwht}},
\code{\link{acqnoatt}},
\code{\link{acqprisn}},
\code{\link{acqrachl}},
\code{\link{acqreps}},
\code{\link{acqshawn}},
\code{\link{acqsocs}},
\code{\link{acqunemp}},
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvalib}},
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhme}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkno}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}},
\code{\link{acqwkwht}},
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{comamind}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{educ1}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{fricall}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndknow}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{grpother}},
\code{\link{grpsprts}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{hlpinflu}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{lentto}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{peocntct}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkedto}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}},
\code{\link{trtasian}},
\code{\link{trtattnd}},
\code{\link{trtblack}},
\code{\link{trtbrnda}},
\code{\link{trtchild}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtcops}},
\code{\link{trtdems}},
\code{\link{trtelecs}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtjose}},
\code{\link{trtkaren}},
\code{\link{trtkeith}},
\code{\link{trtkevin}},
\code{\link{trtlaws}},
\code{\link{trtlib}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtmils}},
\code{\link{trtmyrac}},
\code{\link{trtnoatt}},
\code{\link{trtprisn}},
\code{\link{trtrachl}},
\code{\link{trtreps}},
\code{\link{trtshawn}},
\code{\link{trtsocs}},
\code{\link{trtunemp}},
\code{\link{trtwhite}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{youinflu}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Education}
\concept{Social Networks}
\keyword{variable}
