% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{educ30}
\alias{educ30}
\title{R use education site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
educ30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
C. Other educational site?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} 1-2 times
\item \verb{[3]} 3-5 times
\item \verb{[4]} more than 5 times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2396/vshow}{GSS Data Explorer page for \code{educ30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 290 \tab 193 \tab 105 \tab 79 \tab 75 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 456 \tab 245 \tab 126 \tab 162 \tab 94 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 327 \tab 160 \tab 94 \tab 97 \tab 12 \tab - \tab 2812 \cr
   Total \tab 5878 \tab 1073 \tab 598 \tab 325 \tab 338 \tab 181 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{bookmark}},
\code{\link{catdrcty}},
\code{\link{emaillnk}},
\code{\link{fin30}},
\code{\link{govt30}},
\code{\link{hyperlnk}},
\code{\link{news30}},
\code{\link{schl30}},
\code{\link{work30}}
}
\concept{Information Society}
\keyword{variable}
