% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{educ5}
\alias{educ5}
\title{Education level of person #5}
\source{
General Social Survey https://gss.norc.org
}
\description{
educ5
}
\details{
Question This card (CARD BH) lists general levels of education. As far as you know, what is (NAME)''s highest level of education?
PROBE:  Your best guess. RECORD VERBATIM IF NOT CODEABLE. ASK FOR EACH NAME.
5. Education of Fifth Person
}
\section{Values}{

\itemize{
\item \code{0} 1-6 years
\item \code{1} 7-9 years
\item \code{2} 10-12 years
\item \code{3} h.s. grad
\item \code{4} some college
\item \code{5} asso. degree
\item \code{6} bach. degree
\item \code{7} grad or prof
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/944/vshow}{GSS Data Explorer page for \code{educ5}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1-6 years \tab 7-9 years \tab 10-12 years \tab h.s. grad \tab some college \tab asso. degree \tab bach. degree \tab grad or prof \tab don't know \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 6 \tab 22 \tab 24 \tab 104 \tab 77 \tab 14 \tab 67 \tab 52 \tab 7 \tab 8 \tab 1534 \cr
   2004 \tab 2600 \tab - \tab 3 \tab 6 \tab 35 \tab 26 \tab 17 \tab 47 \tab 28 \tab 4 \tab 46 \tab 2812 \cr
   Total \tab 3753 \tab 6 \tab 25 \tab 30 \tab 139 \tab 103 \tab 31 \tab 114 \tab 80 \tab 11 \tab 54 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}}
}
\concept{Social Networks}
\keyword{variable}
