% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{educop}
\alias{educop}
\title{Equal opportunity to obtain education?}
\source{
General Social Survey https://gss.norc.org
}
\description{
educop
}
\details{
Question A. Does everyone in this country have an opportunity to obtain an education corresponding to their abilities and talents?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/241/vshow}{GSS Data Explorer page for \code{educop}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 1028 \tab 411 \tab 22 \tab 12 \tab 1473 \cr
   Total \tab 0 \tab 1028 \tab 411 \tab 22 \tab 12 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}},
\code{\link{tax}}
}
\concept{Core}
\keyword{variable}
