% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_420.R
\name{eff911a}
\alias{eff911a}
\title{How effective r thinks developing emergency plans}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911a
}
\details{
Question How effective do you think developing emergency plans (evacuation, meeting places) is for people dealing with  terrorism?
Would you say 1, not at all effective, 5, extremely effective, or you may use any number in between?
}
\section{Values}{

\itemize{
\item \code{1} 1 not at all effective
\item \code{2} 2
\item \code{3} 3
\item \code{4} 4
\item \code{5} 5 extremely effective
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3844/vshow}{GSS Data Explorer page for \code{eff911a}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 not at all effective \tab 2 \tab 3 \tab 4 \tab 5 extremely effective \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 167 \tab 135 \tab 430 \tab 197 \tab 351 \tab 67 \tab 5 \tab 2023 \cr
   Total \tab 671 \tab 167 \tab 135 \tab 430 \tab 197 \tab 351 \tab 67 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911a}},
\code{\link{done911b}},
\code{\link{done911c}},
\code{\link{eff911b}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
