% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_421.R
\name{eff911c}
\alias{eff911c}
\title{How effective r thinks purchasing things to make them safer}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911c
}
\details{
Question How effective do you think purchasing things to make them safer (gas masks, duct tape, things to make their house safer, etc.) is for people dealing with terrorism? (Would you say 1, not at all effective, 5, extremely effective, or you may use any number in between?)
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 not at all effective
\item \verb{[2]} 2
\item \verb{[3]} 3
\item \verb{[4]} 4
\item \verb{[5]} 5 extremely effective
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3848/vshow}{GSS Data Explorer page for \code{eff911c}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 not at all effective \tab 2 \tab 3 \tab 4 \tab 5 extremely effective \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 378 \tab 186 \tab 373 \tab 139 \tab 202 \tab 69 \tab 5 \tab 2023 \cr
   Total \tab 671 \tab 378 \tab 186 \tab 373 \tab 139 \tab 202 \tab 69 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911d}},
\code{\link{done911e}},
\code{\link{done911f}},
\code{\link{done911g}},
\code{\link{done911h}},
\code{\link{eff911d}},
\code{\link{eff911e}},
\code{\link{eff911f}},
\code{\link{eff911g}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
