% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_421.R
\name{eff911d}
\alias{eff911d}
\title{How effective r thinks learning where to get info about terrorism}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911d
}
\details{
Question How effective do you think learning where to get more information about terrorism is for people dealing with terrorism? (Would you say 1, not at all effective, 5, extremely effective, or you may use any number in between?)
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 not at all effective
\item \verb{[2]} 2
\item \verb{[3]} 3
\item \verb{[4]} 4
\item \verb{[5]} 5 extremely effective
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3850/vshow}{GSS Data Explorer page for \code{eff911d}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 not at all effective \tab 2 \tab 3 \tab 4 \tab 5 extremely effective \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 671 \tab 231 \tab 128 \tab 352 \tab 216 \tab 343 \tab 75 \tab 7 \tab - \tab 2023 \cr
   Total \tab 671 \tab 231 \tab 128 \tab 352 \tab 216 \tab 343 \tab 75 \tab 7 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911a}},
\code{\link{done911b}},
\code{\link{done911c}},
\code{\link{done911d}},
\code{\link{done911e}},
\code{\link{done911f}},
\code{\link{done911g}},
\code{\link{done911h}},
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911l}},
\code{\link{done911m}},
\code{\link{done911n}},
\code{\link{doneelse}},
\code{\link{eff911a}},
\code{\link{eff911b}},
\code{\link{eff911c}},
\code{\link{eff911e}},
\code{\link{eff911f}},
\code{\link{eff911g}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}},
\code{\link{eff911m}},
\code{\link{eff911n}}

Other Terrorism: 
\code{\link{done911a}},
\code{\link{done911b}},
\code{\link{done911c}},
\code{\link{done911d}},
\code{\link{done911e}},
\code{\link{done911f}},
\code{\link{done911g}},
\code{\link{done911h}},
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911l}},
\code{\link{done911m}},
\code{\link{done911n}},
\code{\link{doneelse}},
\code{\link{eff911a}},
\code{\link{eff911b}},
\code{\link{eff911c}},
\code{\link{eff911e}},
\code{\link{eff911f}},
\code{\link{eff911g}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}},
\code{\link{eff911m}},
\code{\link{eff911n}},
\code{\link{frterror}},
\code{\link{secterr}},
\code{\link{usterror}}
}
\concept{Terrorism}
\concept{Terrorism Preparedness}
\keyword{variable}
