% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_422.R
\name{eff911l}
\alias{eff911l}
\title{How effective r thinks avoiding tall buildings}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911l
}
\details{
Question How effective do you think avoiding tall buildings is for people dealing with terrorism? (Would you say 1, not at all effective, 5, extremely effective, or you may use any number in between?)
}
\section{Values}{

\itemize{
\item \code{1} 1 not at all effective
\item \code{2} 2
\item \code{3} 3
\item \code{4} 4
\item \code{5} 5 extremely effective
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3866/vshow}{GSS Data Explorer page for \code{eff911l}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 not at all effective \tab 2 \tab 3 \tab 4 \tab 5 extremely effective \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 806 \tab 156 \tab 186 \tab 43 \tab 87 \tab 66 \tab 8 \tab 2023 \cr
   Total \tab 671 \tab 806 \tab 156 \tab 186 \tab 43 \tab 87 \tab 66 \tab 8 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911l}},
\code{\link{done911m}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
