% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{effort}
\alias{effort}
\title{In past 30 days feel everything thing effort}
\source{
General Social Survey https://gss.norc.org
}
\description{
effort
}
\details{
Question In the past 30 days, about how often did you feel:
E. That everything was an effort.
}
\section{Values}{

\itemize{
\item \code{1} all of the time
\item \code{2} most of the time
\item \code{3} some of the time
\item \code{4} little of time
\item \code{5} none of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2118/vshow}{GSS Data Explorer page for \code{effort}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all of the time \tab most of the time \tab some of the time \tab little of time \tab none of the time \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 65 \tab 121 \tab 324 \tab 433 \tab 479 \tab 6 \tab 17 \tab 2832 \cr
   Total \tab 1387 \tab 65 \tab 121 \tab 324 \tab 433 \tab 479 \tab 6 \tab 17 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{beausprt}},
\code{\link{fidgety}},
\code{\link{godlove}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{uniongod}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
