% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{eftotlt}
\alias{eftotlt}
\title{Total number of students enrolled in college respondent attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
eftotlt
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
B. Total Fall Enrollment
}
\section{Values}{

\itemize{
\item \code{1} 0-467 students enrolled
\item \code{2} 468 or more students
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/70/vshow}{GSS Data Explorer page for \code{eftotlt}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 0-467 students enrolled \tab 468 or more students \tab no answer \tab Total \cr
   2012 \tab 1347 \tab 14 \tab 610 \tab 3 \tab 1974 \cr
   2014 \tab 1750 \tab 8 \tab 776 \tab 4 \tab 2538 \cr
   2016 \tab 1937 \tab 11 \tab 912 \tab 7 \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 7382 \tab 33 \tab 2298 \tab 14 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
