% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{egomeans}
\alias{egomeans}
\title{Life meaningful only if you provide meaning}
\source{
General Social Survey https://gss.norc.org
}
\description{
egomeans
}
\details{
Question Do you agree or disagree with the following. . .
F. Life is only meaningful if you provide the meaning yourself.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4662/vshow}{GSS Data Explorer page for \code{egomeans}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 280 \tab 476 \tab 186 \tab 192 \tab 108 \tab 49 \tab 68 \tab 1517 \cr
   1998 \tab 1548 \tab 126 \tab 417 \tab 219 \tab 251 \tab 152 \tab 50 \tab 69 \tab 2832 \cr
   2008 \tab 658 \tab 192 \tab 464 \tab 190 \tab 336 \tab 158 \tab 21 \tab 4 \tab 2023 \cr
   2018 \tab 1173 \tab 177 \tab 381 \tab 158 \tab 297 \tab 141 \tab 14 \tab 7 \tab 2348 \cr
   Total \tab 3537 \tab 775 \tab 1738 \tab 753 \tab 1076 \tab 559 \tab 134 \tab 148 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{relexper}},
\code{\link{religkid}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\keyword{variable}
