% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{eldcost}
\alias{eldcost}
\title{Who should pay for help for elderly}
\source{
General Social Survey https://gss.norc.org
}
\description{
eldcost
}
\details{
Question And who do you think should primarily cover the costs of this help to these elderly people?
}
\section{Values}{

\itemize{
\item \code{1} the elderly people themselves or their family
\item \code{2} the government/public funds
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4472/vshow}{GSS Data Explorer page for \code{eldcost}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab the elderly people themselves or their family \tab the government/public funds \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 567 \tab 482 \tab 247 \tab 6 \tab - \tab 1974 \cr
   2022 \tab 2385 \tab 320 \tab 586 \tab 242 \tab 5 \tab 6 \tab 3544 \cr
   Total \tab 3057 \tab 887 \tab 1068 \tab 489 \tab 11 \tab 6 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
