% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{eldersup}
\alias{eldersup}
\title{Adult children are imp to help elderly parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
eldersup
}
\details{
Question To what extent do you agree or disagree with the following statements?
F. Adult children are an important source of help for elderly parents.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4462/vshow}{GSS Data Explorer page for \code{eldersup}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2012 \tab 672 \tab 267 \tab 832 \tab 98 \tab 70 \tab 9 \tab 20 \tab 6 \tab 1974 \cr
   Total \tab 672 \tab 267 \tab 832 \tab 98 \tab 70 \tab 9 \tab 20 \tab 6 \tab 1974 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
