% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{eldfnce}
\alias{eldfnce}
\title{Grandparents should help grandchildren financially}
\source{
General Social Survey https://gss.norc.org
}
\description{
eldfnce
}
\details{
Question Consider an elderly couple. One of their grandchildren is in need of temporary financial help because the company he/she worked for closed down recently. Do you think they should, or should not, try their best to help their grandchild financially? Pleas
}
\section{Values}{

\itemize{
\item \code{1} yes, they definitely should
\item \code{7} no, they definitely should not
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7780/vshow}{GSS Data Explorer page for \code{eldfnce}}.
}

\keyword{variable}
