% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{eldhelp}
\alias{eldhelp}
\title{Who should provide help for elderly}
\source{
General Social Survey https://gss.norc.org
}
\description{
eldhelp
}
\details{
Question Thinking about elderly people who need some help in their everyday lives, such as help with grocery shopping, cleaning the house, doing the laundry etc.
Who do you think should primarily provide this help?
}
\section{Values}{

\itemize{
\item \verb{[1]} family members
\item \verb{[2]} government agencies
\item \verb{[3]} non-profit organizations (e.g., charitable organizations, churches/religious organizations)
\item \verb{[4]} private providers of this kind of help
\item \verb{[5]} employers
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4471/vshow}{GSS Data Explorer page for \code{eldhelp}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab family members \tab government agencies \tab non-profit organizations (e.g., charitable organizations, churches/religious organizations) \tab private providers of this kind of help \tab don't know \tab no answer \tab employers \tab skipped on web \tab not available in this year \tab Total \cr
   2012 \tab 672 \tab 763 \tab 191 \tab 95 \tab 107 \tab 140 \tab 6 \tab - \tab - \tab - \tab 1974 \cr
   2022 \tab 2384 \tab 523 \tab 290 \tab 79 \tab 57 \tab 193 \tab 6 \tab 7 \tab 5 \tab - \tab 3544 \cr
   Total \tab 3056 \tab 1286 \tab 481 \tab 174 \tab 164 \tab 333 \tab 12 \tab 7 \tab 5 \tab 0 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{buythngs}},
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{chldcare}},
\code{\link{clean1}},
\code{\link{cohabfst}},
\code{\link{cohabit}},
\code{\link{cohabok}},
\code{\link{cooking1}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{deckids}},
\code{\link{dinner}},
\code{\link{divbest}},
\code{\link{divhubby}},
\code{\link{divifkd1}},
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokd1}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{eldcost}},
\code{\link{eldersup}},
\code{\link{evdiv}},
\code{\link{fambudgt}},
\code{\link{famsuffr}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{fejobind}},
\code{\link{femarry}},
\code{\link{fewknokd}},
\code{\link{fewrksup}},
\code{\link{fourkids}},
\code{\link{hapifwrk}},
\code{\link{happy7}},
\code{\link{hhclean1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}},
\code{\link{kidsocst}},
\code{\link{kidsuffr}},
\code{\link{kidtrble}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{mardiv}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{mawork14}},
\code{\link{mawrkwrm}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{memarry}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{rfamlook}},
\code{\link{rhhwork}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{sexhar}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{singlpar}},
\code{\link{spborn}},
\code{\link{spevdiv}},
\code{\link{spfalook}},
\code{\link{sphhwork}},
\code{\link{splive}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{teensex1}},
\code{\link{threkids}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhm1}},
\code{\link{tiredhme}},
\code{\link{tiredwk1}},
\code{\link{tiredwrk}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{twokids}},
\code{\link{weekend}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}

Other Aged: 
\code{\link{aged}},
\code{\link{aidold}},
\code{\link{eldcost}},
\code{\link{natsoc}},
\code{\link{spretire}}
}
\concept{Aged}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
