% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{elecfair}
\alias{elecfair}
\title{How fair last natl election: opprtunities of candidate}
\source{
General Social Survey https://gss.norc.org
}
\description{
elecfair
}
\details{
Question Thinking of the last national election in America, how fair was it regarding the opportunities of the candidates and parties to campaign?
}
\section{Values}{

\itemize{
\item \code{1} very fair
\item \code{2} somewhat fair
\item \code{3} neither fair nor unfair
\item \code{4} somewhat unfair
\item \code{5} very unfair
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4956/vshow}{GSS Data Explorer page for \code{elecfair}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very fair \tab somewhat fair \tab neither fair nor unfair \tab somewhat unfair \tab very unfair \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 521 \tab 479 \tab 221 \tab 145 \tab 60 \tab 40 \tab 6 \tab 2812 \cr
   2014 \tab 1274 \tab 358 \tab 433 \tab 176 \tab 100 \tab 46 \tab 145 \tab 6 \tab 2538 \cr
   Total \tab 2614 \tab 879 \tab 912 \tab 397 \tab 245 \tab 106 \tab 185 \tab 12 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
