% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{elecvote}
\alias{elecvote}
\title{How honest last natl election: counting of votes}
\source{
General Social Survey https://gss.norc.org
}
\description{
elecvote
}
\details{
Question Thinking of the last national election in America, how honest was it regarding the counting and reporting of the votes?
}
\section{Values}{

\itemize{
\item \code{1} very honest
\item \code{2} somewhat honest
\item \code{3} neither honest nor dishonest
\item \code{4} somewhat dishonest
\item \code{5} very dishonest
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4955/vshow}{GSS Data Explorer page for \code{elecvote}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very honest \tab somewhat honest \tab neither honest nor dishonest \tab somewhat dishonest \tab very dishonest \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 195 \tab 364 \tab 206 \tab 339 \tab 328 \tab 34 \tab 6 \tab 2812 \cr
   2014 \tab 1274 \tab 305 \tab 422 \tab 167 \tab 126 \tab 52 \tab 186 \tab 6 \tab 2538 \cr
   Total \tab 2614 \tab 500 \tab 786 \tab 373 \tab 465 \tab 380 \tab 220 \tab 12 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
