% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{emailhr}
\alias{emailhr}
\title{Email hours per week}
\source{
General Social Survey https://gss.norc.org
}
\description{
emailhr
}
\details{
Question About how many minutes or hours per week do you spend sending and answering electronic mail or e-mail?
}
\section{Values}{

\itemize{
\item \verb{[0]} 0 hours
\item \verb{[168]} 168 hours
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2382/vshow}{GSS Data Explorer page for \code{emailhr}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 hours \tab 1 \tab 10 \tab 12 \tab 14 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 21 \tab 24 \tab 25 \tab 27 \tab 3 \tab 30 \tab 35 \tab 4 \tab 40 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab 11 \tab 13 \tab 22 \tab 32 \tab 41 \tab 42 \tab 56 \tab 70 \tab 28 \tab 36 \tab 38 \tab 43 \tab 45 \tab 48 \tab 55 \tab 90 \tab 100 \tab 168 hours \tab 33 \tab 34 \tab 65 \tab 72 \tab 80 \tab 84 \tab 126 \tab 150 \tab 17 \tab 23 \tab 99 \tab 44 \tab 47 \tab 75 \tab 37 \tab 124 \tab skipped on web \tab 19 \tab 26 \tab 51 \tab 54 \tab 63 \tab 91 \tab not available in this year \tab Total \cr
   2000 \tab 1424 \tab 538 \tab 205 \tab 53 \tab 6 \tab 7 \tab 25 \tab 3 \tab 2 \tab 138 \tab 21 \tab 2 \tab 1 \tab 3 \tab 1 \tab 98 \tab 5 \tab 2 \tab 48 \tab 10 \tab 71 \tab 2 \tab 20 \tab 1 \tab 35 \tab 23 \tab 1 \tab 4 \tab 68 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 855 \tab 617 \tab 301 \tab 90 \tab 18 \tab 14 \tab 27 \tab 3 \tab 2 \tab 241 \tab 36 \tab 3 \tab 5 \tab 11 \tab - \tab 123 \tab 22 \tab 2 \tab 67 \tab 6 \tab 129 \tab 5 \tab 47 \tab 3 \tab 62 \tab 30 \tab 3 \tab 6 \tab 23 \tab 4 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2178 \tab 109 \tab 125 \tab 41 \tab 7 \tab 3 \tab 12 \tab - \tab 1 \tab 89 \tab 18 \tab 3 \tab 1 \tab 7 \tab - \tab 29 \tab 9 \tab 4 \tab 26 \tab 4 \tab 62 \tab 5 \tab 13 \tab - \tab 24 \tab 12 \tab 3 \tab 5 \tab 12 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2010 \tab 930 \tab 246 \tab 211 \tab 64 \tab 10 \tab 10 \tab 27 \tab 3 \tab 2 \tab 146 \tab 43 \tab 4 \tab 1 \tab 15 \tab - \tab 69 \tab 22 \tab 3 \tab 48 \tab 9 \tab 59 \tab 2 \tab 21 \tab 3 \tab 35 \tab 29 \tab 4 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 917 \tab 287 \tab 186 \tab 51 \tab 12 \tab 13 \tab 23 \tab 3 \tab 2 \tab 106 \tab 42 \tab 3 \tab 2 \tab 13 \tab - \tab 59 \tab 15 \tab 3 \tab 35 \tab 14 \tab 57 \tab 5 \tab 31 \tab 6 \tab 34 \tab 24 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1060 \tab 379 \tab 295 \tab 84 \tab 10 \tab 18 \tab 34 \tab 3 \tab - \tab 152 \tab 59 \tab 2 \tab 5 \tab 20 \tab 2 \tab 81 \tab 37 \tab 6 \tab 52 \tab 12 \tab 77 \tab 4 \tab 28 \tab 3 \tab 57 \tab 26 \tab 3 \tab 5 \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1205 \tab 416 \tab 306 \tab 100 \tab 27 \tab 9 \tab 36 \tab 3 \tab 1 \tab 183 \tab 76 \tab 5 \tab 1 \tab 24 \tab - \tab 77 \tab 33 \tab 8 \tab 62 \tab 25 \tab 92 \tab 13 \tab 27 \tab 7 \tab 44 \tab 42 \tab 4 \tab 7 \tab 6 \tab - \tab - \tab 5 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 925 \tab 356 \tab 247 \tab 93 \tab 15 \tab 22 \tab 33 \tab 3 \tab 1 \tab 155 \tab 61 \tab 2 \tab 7 \tab 24 \tab - \tab 78 \tab 33 \tab 7 \tab 52 \tab 24 \tab 66 \tab 9 \tab 35 \tab 3 \tab 45 \tab 22 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 5 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1384 \tab 248 \tab 471 \tab 160 \tab 30 \tab 25 \tab 62 \tab 7 \tab 5 \tab 355 \tab 93 \tab 5 \tab 4 \tab 26 \tab - \tab 163 \tab 38 \tab 10 \tab 158 \tab 46 \tab 135 \tab 11 \tab 56 \tab 5 \tab 86 \tab 86 \tab 13 \tab 1 \tab 59 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 266 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1271 \tab 524 \tab 408 \tab 124 \tab 22 \tab 16 \tab 38 \tab 2 \tab 5 \tab 247 \tab 72 \tab 3 \tab 4 \tab 34 \tab 1 \tab 144 \tab 34 \tab 6 \tab 99 \tab 59 \tab 141 \tab 11 \tab 52 \tab 7 \tab 66 \tab 62 \tab 10 \tab 28 \tab 21 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3544 \cr
   Total \tab 12149 \tab 3720 \tab 2755 \tab 860 \tab 157 \tab 137 \tab 317 \tab 30 \tab 21 \tab 1812 \tab 521 \tab 32 \tab 31 \tab 177 \tab 4 \tab 921 \tab 248 \tab 51 \tab 647 \tab 209 \tab 889 \tab 67 \tab 330 \tab 38 \tab 488 \tab 356 \tab 43 \tab 70 \tab 196 \tab 14 \tab 11 \tab 11 \tab 8 \tab 4 \tab 5 \tab 4 \tab 12 \tab 27 \tab 5 \tab 3 \tab 1 \tab 21 \tab 12 \tab 4 \tab 5 \tab 6 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 8 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 266 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 0 \tab 27741 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artgn}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyinfgn}},
\code{\link{buyit12}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{gamegn}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{huclean}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{inperson}},
\code{\link{intaccss}},
\code{\link{invest12}},
\code{\link{investgn}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobhrw}},
\code{\link{jobinfo}},
\code{\link{jobminw}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{letters}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{meetings}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{newjobgn}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numcntct}},
\code{\link{numhome}},
\code{\link{numprobs}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{peoplegn}},
\code{\link{perfrmgn}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polcamp}},
\code{\link{polcangn}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldisgn}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polfrel}},
\code{\link{polinf12}},
\code{\link{polinfgn}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{polpapr}},
\code{\link{poltough}},
\code{\link{poltv}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{travelgn}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}

Other Computers And Internet: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{autonojb}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{intaccss}},
\code{\link{internet}},
\code{\link{intrhome}},
\code{\link{invest12}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numhome}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polinf12}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{poltough}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
