% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{emailyr}
\alias{emailyr}
\title{When r first began using email}
\source{
General Social Survey https://gss.norc.org
}
\description{
emailyr
}
\details{
Question A. When did you first begin using email?
}
\section{Values}{

\itemize{
\item \verb{[1]} this year (2002)
\item \verb{[2]} in the last year (2001)
\item \verb{[3]} 1-2 years ago (2000)
\item \verb{[4]} 2-3 years ago (1999)
\item \verb{[5]} 3-4 years ago (1998)
\item \verb{[6]} 4-5 years ago (1997)
\item \verb{[7]} 5 or more years ago (1990-1996)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2619/vshow}{GSS Data Explorer page for \code{emailyr}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab this year (2002) \tab in the last year (2001) \tab 1-2 years ago (2000) \tab 2-3 years ago (1999) \tab 3-4 years ago (1998) \tab 4-5 years ago (1997) \tab 5 or more years ago (1990-1996) \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1271 \tab 43 \tab 114 \tab 184 \tab 219 \tab 194 \tab 189 \tab 534 \tab 4 \tab 13 \tab 2765 \cr
   Total \tab 1271 \tab 43 \tab 114 \tab 184 \tab 219 \tab 194 \tab 189 \tab 534 \tab 4 \tab 13 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webable}},
\code{\link{webyr}}
}
\concept{Information Society}
\keyword{variable}
