% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_285.R
\name{emfamloc}
\alias{emfamloc}
\title{R send or receive to or from family members at}
\source{
General Social Survey https://gss.norc.org
}
\description{
emfamloc
}
\details{
Question Thinking now about all your home, work or other computers,
G. Family members who live with you?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2601/vshow}{GSS Data Explorer page for \code{emfamloc}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2423 \tab 50 \tab 284 \tab 60 \tab 2817 \cr
   2002 \tab 2167 \tab 134 \tab 423 \tab 41 \tab 2765 \cr
   Total \tab 4590 \tab 184 \tab 707 \tab 101 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emchurch}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emgroups}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}}
}
\concept{Information Society}
\keyword{variable}
