% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_285.R
\name{emfri}
\alias{emfri}
\title{R send or receive to or from friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
emfri
}
\details{
Question Thinking now about all your home, work or other computers,
I. Friends?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2603/vshow}{GSS Data Explorer page for \code{emfri}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2423 \tab 265 \tab 72 \tab 57 \tab 2817 \cr
   2002 \tab 2167 \tab 504 \tab 53 \tab 41 \tab 2765 \cr
   Total \tab 4590 \tab 769 \tab 125 \tab 98 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emgroups}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}}
}
\concept{Information Society}
\keyword{variable}
