% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{emminh}
\alias{emminh}
\title{Minutes per week use computer email}
\source{
General Social Survey https://gss.norc.org
}
\description{
emminh
}
\details{
Question About how many minutes or hours a week do you use your home computer for each of the following:
D. Sending or receiving e-mail?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2562/vshow}{GSS Data Explorer page for \code{emminh}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 15 \tab 2 \tab 20 \tab 3 \tab 30 \tab 40 \tab 45 \tab 5 \tab 50 \tab don't know \tab no answer \tab 12 \tab 14 \tab 25 \tab 35 \tab 8 \tab Total \cr
   2000 \tab 2704 \tab 3 \tab 9 \tab 12 \tab 1 \tab 10 \tab 1 \tab 35 \tab 1 \tab 4 \tab 5 \tab 3 \tab 2 \tab 27 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2413 \tab 9 \tab 41 \tab 46 \tab 1 \tab 31 \tab 2 \tab 124 \tab 4 \tab 7 \tab 30 \tab 4 \tab 3 \tab 44 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2765 \cr
   Total \tab 5117 \tab 12 \tab 50 \tab 58 \tab 2 \tab 41 \tab 3 \tab 159 \tab 5 \tab 11 \tab 35 \tab 7 \tab 5 \tab 71 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compwork}},
\code{\link{emhrh}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{perhrh}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwhrh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
