% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_284.R
\name{emmino}
\alias{emmino}
\title{Minutes per week using other computer email}
\source{
General Social Survey https://gss.norc.org
}
\description{
emmino
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
D. Sending or receiving e-mail?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2591/vshow}{GSS Data Explorer page for \code{emmino}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 10 \tab 15 \tab 2 \tab 20 \tab 30 \tab 45 \tab 5 \tab no answer \tab 1 \tab 25 \tab 6 \tab don't know \tab Total \cr
   2000 \tab 2756 \tab 5 \tab 4 \tab 3 \tab 2 \tab 10 \tab 1 \tab 3 \tab 33 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2626 \tab 19 \tab 9 \tab 1 \tab 7 \tab 42 \tab 1 \tab 7 \tab 44 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2765 \cr
   Total \tab 5382 \tab 24 \tab 13 \tab 4 \tab 9 \tab 52 \tab 2 \tab 10 \tab 77 \tab 3 \tab 1 \tab 1 \tab 4 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emhro}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{wrkhro}},
\code{\link{wrkmino}},
\code{\link{wwwhro}},
\code{\link{wwwmino}}
}
\concept{Information Society}
\keyword{variable}
