% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{emminw}
\alias{emminw}
\title{Minutes per week using work computer email}
\source{
General Social Survey https://gss.norc.org
}
\description{
emminw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
E. Sending or receiving e-mail?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2575/vshow}{GSS Data Explorer page for \code{emminw}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 15 \tab 2 \tab 20 \tab 30 \tab 35 \tab 45 \tab 5 \tab 50 \tab don't know \tab no answer \tab 3 \tab 40 \tab 6 \tab 7 \tab Total \cr
   2000 \tab 2719 \tab 1 \tab 3 \tab 4 \tab 5 \tab 4 \tab 27 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 48 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2562 \tab 4 \tab 22 \tab 25 \tab 4 \tab 8 \tab 67 \tab 3 \tab 7 \tab 7 \tab 3 \tab 8 \tab 36 \tab 2 \tab 4 \tab 2 \tab 1 \tab 2765 \cr
   Total \tab 5281 \tab 5 \tab 25 \tab 29 \tab 9 \tab 12 \tab 94 \tab 4 \tab 8 \tab 9 \tab 4 \tab 9 \tab 84 \tab 2 \tab 4 \tab 2 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{ecomminw}},
\code{\link{emhrw}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{perhrw}},
\code{\link{perminw}},
\code{\link{wwwhrw}},
\code{\link{wwwminw}}
}
\concept{Information Society}
\keyword{variable}
