% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{emonitor1}
\alias{emonitor1}
\title{Should government be able to monitor online communication}
\source{
General Social Survey https://gss.norc.org
}
\description{
emonitor1
}
\details{
Question Do you think that the American government should or should not have the right to do the following?… Monitor e-mails, social media content and any other personal information exchanged on the Internet?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should have the right
\item \verb{[2]} probably should have the right
\item \verb{[3]} probably should not have the right
\item \verb{[4]} definitely should not have the right
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8111/vshow}{GSS Data Explorer page for \code{emonitor1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab definitely should have the right \tab probably should have the right \tab probably should not have the right \tab definitely should not have the right \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 90 \tab 273 \tab 463 \tab 654 \tab 138 \tab 1673 \tab 6 \tab 12 \tab 3309 \cr
   Total \tab 0 \tab 90 \tab 273 \tab 463 \tab 654 \tab 138 \tab 1673 \tab 6 \tab 12 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{dataprot}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{modeorg}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
