% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{emoprobs}
\alias{emoprobs}
\title{In the past 7 days, how often r has been bothered by emotional problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
emoprobs
}
\details{
Question In the past seven days, how often have you been bothered by emotional problems such as feeling anxious, depressed or irritable?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} rarely
\item \code{3} sometimes
\item \code{4} often
\item \code{5} always
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7169/vshow}{GSS Data Explorer page for \code{emoprobs}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab rarely \tab sometimes \tab often \tab always \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 727 \tab 686 \tab 601 \tab 234 \tab 81 \tab 2 \tab 17 \tab - \tab 2348 \cr
   2021 \tab 389 \tab 765 \tab 1092 \tab 1126 \tab 478 \tab 153 \tab - \tab 1 \tab 28 \tab 4032 \cr
   Total \tab 389 \tab 1492 \tab 1778 \tab 1727 \tab 712 \tab 234 \tab 2 \tab 18 \tab 28 \tab 6380 \cr
}
}

\keyword{variable}
