% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_299.R
\name{emotqol}
\alias{emotqol}
\title{Treatment for emotion improve quality of li}
\source{
General Social Survey https://gss.norc.org
}
\description{
emotqol
}
\details{
Question Say you went to treatment for an EMOTIONAL HEALTH problem such as feeling depressed or anxious that affects your work and other daily activities so that you accomplish less than you would like...
A. Would you definitely expect, probably expect, probably not expect, or definitely not expect your overall quality of life to improve?
}
\section{Values}{

\itemize{
\item \code{1} definitely expect
\item \code{2} probably expect
\item \code{3} probably not expect
\item \code{4} definitely not expect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2720/vshow}{GSS Data Explorer page for \code{emotqol}}.
}

\seealso{
Other Health Status: 
\code{\link{downfeel}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}}

Other Health: 
\code{\link{downfeel}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}}
}
\concept{Health}
\concept{Health Status}
\keyword{variable}
