% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{emotrely}
\alias{emotrely}
\title{Treatment for emotion rely less on others}
\source{
General Social Survey https://gss.norc.org
}
\description{
emotrely
}
\details{
Question Say you went to treatment for an EMOTIONAL HEALTH problem such as feeling depressed or anxious that affects your work and other daily activities so that you accomplish less than you would like...
D. Would you definitely expect, probably expect, probably not expect, or definitely not expect to need to rely less on others?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely expect
\item \verb{[2]} probably expect
\item \verb{[3]} probably not expect
\item \verb{[4]} definitely not expect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2723/vshow}{GSS Data Explorer page for \code{emotrely}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely expect \tab probably expect \tab probably not expect \tab definitely not expect \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2331 \tab 103 \tab 186 \tab 88 \tab 29 \tab 44 \tab 36 \tab 2817 \cr
   Total \tab 2331 \tab 103 \tab 186 \tab 88 \tab 29 \tab 44 \tab 36 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{emotfeel}}
}
\concept{Health Status}
\keyword{variable}
