% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{empathy3}
\alias{empathy3}
\title{R feels protective towards people taken advantage}
\source{
General Social Survey https://gss.norc.org
}
\description{
empathy3
}
\details{
Question The following statements ask about your thoughts and feelings in various situations. For each item indicate how well it describes you by choosing the number on the showcard where 1 indicates that it does not describe you very well  and 5 means that it does describe you very well. Of course numbers 2-4 indicate that how well it describes you are In between these points.
C. When I see someone being taken advantage of, I feel kind of protective towards them.
}
\section{Values}{

\itemize{
\item \verb{[1]} does not describe very well
\item \verb{[5]} describes very well
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2872/vshow}{GSS Data Explorer page for \code{empathy3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab does not describe very well \tab 2 \tab 3 \tab 4 \tab describes very well \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 58 \tab 60 \tab 168 \tab 422 \tab 644 \tab 14 \tab 6 \tab 2765 \cr
   2004 \tab 1472 \tab 31 \tab 43 \tab 144 \tab 459 \tab 653 \tab 5 \tab 5 \tab 2812 \cr
   Total \tab 2865 \tab 89 \tab 103 \tab 312 \tab 881 \tab 1297 \tab 19 \tab 11 \tab 5577 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{empathy1}},
\code{\link{empathy2}},
\code{\link{empathy4}}
}
\concept{Altruism}
\keyword{variable}
