% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_323.R
\name{empathy6}
\alias{empathy6}
\title{R often quite touched by things}
\source{
General Social Survey https://gss.norc.org
}
\description{
empathy6
}
\details{
Question The following statements ask about your thoughts and feelings in various situations. For each item indicate how well it describes you by choosing the number on the showcard where 1 indicates that it does not describe you very well  and 5 means that it does describe you very well. Of course numbers 2-4 indicate that how well it describes you are In between these points.
F. I am often quite touched by things that I see happen.
}
\section{Values}{

\itemize{
\item \code{1} does not describe very well
\item \code{5} describes very well
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2875/vshow}{GSS Data Explorer page for \code{empathy6}}.
}

\seealso{
Other Altruism: 
\code{\link{cutahead}},
\code{\link{empathy5}},
\code{\link{empathy7}},
\code{\link{givblood}},
\code{\link{givchrty}},
\code{\link{givhmlss}},
\code{\link{givseat}},
\code{\link{retchnge}},
\code{\link{volchrty}}

Other Feelings: 
\code{\link{empathy5}},
\code{\link{empathy7}}
}
\concept{Altruism}
\concept{Feelings}
\keyword{variable}
