% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_323.R
\name{empathy7}
\alias{empathy7}
\title{R describes oneself as a soft-hearted person}
\source{
General Social Survey https://gss.norc.org
}
\description{
empathy7
}
\details{
Question The following statements ask about your thoughts and feelings in various situations. For each item indicate how well it describes you by choosing the number on the showcard where 1 indicates that it does not describe you very well  and 5 means that it does describe you very well. Of course numbers 2-4 indicate that how well it describes you are In between these points.
G. I would describe myself as a pretty soft-hearted person.
}
\section{Values}{

\itemize{
\item \verb{[1]} does not describe very well
\item \verb{[5]} describes very well
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2876/vshow}{GSS Data Explorer page for \code{empathy7}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab does not describe very well \tab 2 \tab 3 \tab 4 \tab describes very well \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 50 \tab 69 \tab 234 \tab 337 \tab 666 \tab 10 \tab 6 \tab 2765 \cr
   2004 \tab 1472 \tab 26 \tab 71 \tab 200 \tab 412 \tab 621 \tab 6 \tab 4 \tab 2812 \cr
   Total \tab 2865 \tab 76 \tab 140 \tab 434 \tab 749 \tab 1287 \tab 16 \tab 10 \tab 5577 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{cutahead}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{givblood}},
\code{\link{givchrty}},
\code{\link{givhmlss}},
\code{\link{givseat}},
\code{\link{retchnge}},
\code{\link{volchrty}}
}
\concept{Altruism}
\keyword{variable}
