% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{emphplan}
\alias{emphplan}
\title{Name of insurance or hmo}
\source{
General Social Survey https://gss.norc.org
}
\description{
emphplan
}
\details{
Question Name of Insurance or HMO.
}
\section{Values}{

\itemize{
\item \code{1} aetna
\item \code{2} anthem/wellpoint
\item \code{3} blue cross/blue shield
\item \code{4} blue cross
\item \code{5} blue shield
\item \code{6} cigna
\item \code{7} covington
\item \code{8} harvard pilgrim
\item \code{9} humana
\item \code{10} kaiser permanente
\item \code{11} tufts
\item \code{12} united
\item \code{13} wellpoint
\item \code{14} oxford
\item \code{15} healthnet
\item \code{16} united health care
\item \code{17} health partners
\item \code{18} great west
\item \code{19} unicare
\item \code{20} medcost
\item \code{21} pacific care
\item \code{22} fallon hmo
\item \code{23} arise
\item \code{24} medical mutual of ohio
\item \code{25} florida health care
\item \code{26} hawaiian medical system association
\item \code{27} coventry
\item \code{28} winhealth
\item \code{29} health choice
\item \code{30} preferred care
\item \code{31} mvp
\item \code{32} first health
\item \code{33} umr
\item \code{34} affinity health care
\item \code{35} priority health
\item \code{36} ghi
\item \code{87} christian brothers employee benefit trust
\item \code{89} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3904/vshow}{GSS Data Explorer page for \code{emphplan}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab aetna \tab kaiser permanente \tab tufts \tab united \tab wellpoint \tab oxford \tab healthnet \tab united health care \tab health partners \tab great west \tab unicare \tab anthem/wellpoint \tab medcost \tab pacific care \tab fallon hmo \tab arise \tab medical mutual of ohio \tab florida health care \tab hawaiian medical system association \tab coventry \tab winhealth \tab health choice \tab blue cross/blue shield \tab preferred care \tab mvp \tab first health \tab umr \tab affinity health care \tab priority health \tab ghi \tab blue cross \tab 40 \tab blue shield \tab cigna \tab covington \tab harvard pilgrim \tab other \tab humana \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1371 \tab 49 \tab 22 \tab 2 \tab 48 \tab 1 \tab 17 \tab 3 \tab 8 \tab 3 \tab 3 \tab 2 \tab 22 \tab 5 \tab 2 \tab 4 \tab 2 \tab 7 \tab 2 \tab 6 \tab 6 \tab 5 \tab 2 \tab 215 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 34 \tab 1 \tab 12 \tab 38 \tab 1 \tab 3 \tab 60 \tab 15 \tab 29 \tab 8 \tab 2023 \cr
   Total \tab 1371 \tab 49 \tab 22 \tab 2 \tab 48 \tab 1 \tab 17 \tab 3 \tab 8 \tab 3 \tab 3 \tab 2 \tab 22 \tab 5 \tab 2 \tab 4 \tab 2 \tab 7 \tab 2 \tab 6 \tab 6 \tab 5 \tab 2 \tab 215 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 34 \tab 1 \tab 12 \tab 38 \tab 1 \tab 3 \tab 60 \tab 15 \tab 29 \tab 8 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{natdeg}},
\code{\link{othplan}},
\code{\link{typevisa}},
\code{\link{visa}}
}
\concept{Global Economics}
\keyword{variable}
