% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{empinput}
\alias{empinput}
\title{R involved in any task force for decision-making}
\source{
General Social Survey https://gss.norc.org
}
\description{
empinput
}
\details{
Question Some companies have organized workplace decision-making in ways to get more employee input and involvement. Are you personally involved in any group,  team, committee, or task force that  addresses issues such as product quality, cost cutting, productivity, health and safety, or other workplace issues?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3557/vshow}{GSS Data Explorer page for \code{empinput}}.
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{wealth}}

Other Work: 
\code{\link{compwage}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\concept{Work}
\keyword{variable}
