% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{empself}
\alias{empself}
\title{Does r prefer self-employment or not ?}
\source{
General Social Survey https://gss.norc.org
}
\description{
empself
}
\details{
Question Suppose you were working and could choose between different kinds of jobs. Which of the following would you personally choose?  I would advise...
Please CIRCLE ONE CODE FOR EACH.
A. Being an employee (or) being self-employed?
}
\section{Values}{

\itemize{
\item \verb{[1]} employee
\item \verb{[2]} self-employ
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4524/vshow}{GSS Data Explorer page for \code{empself}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab employee \tab self-employ \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 475 \tab 808 \tab 133 \tab 37 \tab 1537 \cr
   1998 \tab 1609 \tab 313 \tab 758 \tab 128 \tab 24 \tab 2832 \cr
   2006 \tab 2992 \tab 569 \tab 930 \tab 17 \tab 2 \tab 4510 \cr
   Total \tab 4685 \tab 1357 \tab 2496 \tab 278 \tab 63 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
