% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_284.R
\name{emsentp}
\alias{emsentp}
\title{How many personal email messages sent per day}
\source{
General Social Survey https://gss.norc.org
}
\description{
emsentp
}
\details{
Question Thinking now about all your home, work or other computers,
B. About how many of these messages are personal one-to-one, rather than sent to members of a group or e-list?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2596/vshow}{GSS Data Explorer page for \code{emsentp}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 13 \tab 15 \tab 2 \tab 20 \tab 23 \tab 25 \tab 3 \tab 30 \tab 35 \tab 4 \tab 40 \tab 5 \tab 50 \tab 52 \tab 6 \tab 7 \tab 70 \tab 8 \tab 9 \tab 92 \tab don't know \tab no answer \tab 100 \tab 11 \tab 125 \tab 160 \tab 18 \tab 200 \tab 22 \tab 24 \tab 45 \tab 60 \tab not available in this year \tab Total \cr
   2000 \tab 2423 \tab 42 \tab 76 \tab 22 \tab 2 \tab 1 \tab 9 \tab 55 \tab 4 \tab 1 \tab 4 \tab 44 \tab 3 \tab 2 \tab 17 \tab 3 \tab 35 \tab 2 \tab 1 \tab 6 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 54 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2161 \tab 35 \tab 114 \tab 38 \tab 6 \tab - \tab 12 \tab 88 \tab 19 \tab - \tab 5 \tab 69 \tab 7 \tab 1 \tab 38 \tab 3 \tab 60 \tab 5 \tab - \tab 13 \tab 9 \tab 1 \tab 11 \tab 4 \tab - \tab - \tab 41 \tab 2 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab - \tab 2765 \cr
   Total \tab 4584 \tab 77 \tab 190 \tab 60 \tab 8 \tab 1 \tab 21 \tab 143 \tab 23 \tab 1 \tab 9 \tab 113 \tab 10 \tab 3 \tab 55 \tab 6 \tab 95 \tab 7 \tab 1 \tab 19 \tab 12 \tab 2 \tab 14 \tab 6 \tab 1 \tab 1 \tab 95 \tab 2 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 0 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artgn}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyinfgn}},
\code{\link{buyit12}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{gamegn}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{huclean}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{inperson}},
\code{\link{intaccss}},
\code{\link{invest12}},
\code{\link{investgn}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobhrw}},
\code{\link{jobinfo}},
\code{\link{jobminw}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{letters}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{meetings}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{newjobgn}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numcntct}},
\code{\link{numhome}},
\code{\link{numprobs}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{peoplegn}},
\code{\link{perfrmgn}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polcamp}},
\code{\link{polcangn}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldisgn}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polfrel}},
\code{\link{polinf12}},
\code{\link{polinfgn}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{polpapr}},
\code{\link{poltough}},
\code{\link{poltv}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{travelgn}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}

Other Computers And Internet: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{autonojb}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{intaccss}},
\code{\link{internet}},
\code{\link{intrhome}},
\code{\link{invest12}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numhome}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polinf12}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{poltough}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
