% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{engagrnk}
\alias{engagrnk}
\title{Engineers agree nuclear power}
\source{
General Social Survey https://gss.norc.org
}
\description{
engagrnk
}
\details{
Question Please look at Card E14. (On a scale of 1 to 5, where 1 means "Near complete agreement" and 5 means No agreement at all) to what extent Do nuclear engineers agree on the risks and benefits of nuclear power to generate electricity?
}
\section{Values}{

\itemize{
\item \verb{[1]} near complete agreement
\item \verb{[5]} no agreement at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3491/vshow}{GSS Data Explorer page for \code{engagrnk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab near complete agreement \tab 2 \tab 3 \tab 4 \tab no agreement at all \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1578 \tab 89 \tab 151 \tab 127 \tab 35 \tab 10 \tab 44 \tab 10 \tab 2044 \cr
   Total \tab 1578 \tab 89 \tab 151 \tab 127 \tab 35 \tab 10 \tab 44 \tab 10 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstnk}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{nukeelec}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
