% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{engbstnk}
\alias{engbstnk}
\title{Nuclear engineers best for country}
\source{
General Social Survey https://gss.norc.org
}
\description{
engbstnk
}
\details{
Question Please look at Card E16. When making policy recommendations about nuclear power, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests? FI INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
A. Nuclear Engineers
}
\section{Values}{

\itemize{
\item \verb{[1]} what is best for the country
\item \verb{[5]} own narrow interests
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3495/vshow}{GSS Data Explorer page for \code{engbstnk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1578 \tab 109 \tab 133 \tab 115 \tab 44 \tab 31 \tab 27 \tab 7 \tab 2044 \cr
   Total \tab 1578 \tab 109 \tab 133 \tab 115 \tab 44 \tab 31 \tab 27 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstnk}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{engagrnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{nukeelec}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
