% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_388.R
\name{engda}
\alias{engda}
\title{Happy if daughter engineer}
\source{
General Social Survey https://gss.norc.org
}
\description{
engda
}
\details{
Question A. If you had a daughter, how would you feel if she wanted to be an engineer—would you feel happy, unhappy, or would you not care one way or the other?
}
\section{Values}{

\itemize{
\item \code{1} happy
\item \code{2} not care
\item \code{3} unhappy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3522/vshow}{GSS Data Explorer page for \code{engda}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
