% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_388.R
\name{engdo}
\alias{engdo}
\title{Know what engineers do}
\source{
General Social Survey https://gss.norc.org
}
\description{
engdo
}
\details{
Question D. Next, I'd like to ask how you rate your knowledge of what engineers do day-to-day in their jobs. would you say your knowledge of what engineers do day-to-day on their jobs is Excellent, Good, Fair, Poor, Very Poor?
}
\section{Values}{

\itemize{
\item \verb{[1]} excellent
\item \verb{[2]} good
\item \verb{[3]} fair
\item \verb{[4]} poor
\item \verb{[5]} very poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3525/vshow}{GSS Data Explorer page for \code{engdo}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab excellent \tab good \tab fair \tab poor \tab very poor \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 54 \tab 142 \tab 158 \tab 79 \tab 36 \tab 13 \tab 3 \tab 1974 \cr
   Total \tab 1489 \tab 54 \tab 142 \tab 158 \tab 79 \tab 36 \tab 13 \tab 3 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
