% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_389.R
\name{engearn}
\alias{engearn}
\title{Engineers earn less}
\source{
General Social Survey https://gss.norc.org
}
\description{
engearn
}
\details{
Question J. Engineers earn less than other people with equally demanding jobs. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3535/vshow}{GSS Data Explorer page for \code{engearn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 5 \tab 33 \tab 333 \tab 39 \tab 70 \tab 5 \tab 1974 \cr
   Total \tab 1489 \tab 5 \tab 33 \tab 333 \tab 39 \tab 70 \tab 5 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engfun}},
\code{\link{engint}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{farming}},
\code{\link{journlsm}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
