% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{england}
\alias{england}
\title{Liking for england}
\source{
General Social Survey https://gss.norc.org
}
\description{
england
}
\details{
Question You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? READ EACH ITEM.
C. England
}
\section{Values}{

\itemize{
\item \code{0} +5
\item \code{1} +4
\item \code{2} +3
\item \code{3} +2
\item \code{4} +1
\item \code{5} -1
\item \code{6} -2
\item \code{7} -3
\item \code{8} -4
\item \code{9} -5
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974, 1975, 1977, 1982, 1983, 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5492/vshow}{GSS Data Explorer page for \code{england}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab +5 \tab +4 \tab +3 \tab +2 \tab +1 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab iap \tab Total \cr
   1974 \tab - \tab 250 \tab 294 \tab 358 \tab 202 \tab 153 \tab 62 \tab 25 \tab 24 \tab 14 \tab 37 \tab 65 \tab - \tab 1484 \cr
   1975 \tab - \tab 244 \tab 241 \tab 367 \tab 198 \tab 200 \tab 57 \tab 24 \tab 32 \tab 13 \tab 24 \tab 90 \tab - \tab 1490 \cr
   1977 \tab - \tab 276 \tab 260 \tab 326 \tab 193 \tab 197 \tab 60 \tab 33 \tab 20 \tab 13 \tab 28 \tab 124 \tab - \tab 1530 \cr
   1982 \tab - \tab 349 \tab 329 \tab 410 \tab 256 \tab 211 \tab 63 \tab 34 \tab 32 \tab 22 \tab 25 \tab 129 \tab - \tab 1860 \cr
   1983 \tab - \tab 331 \tab 304 \tab 350 \tab 208 \tab 205 \tab 41 \tab 26 \tab 28 \tab 15 \tab 20 \tab 71 \tab - \tab 1599 \cr
   1985 \tab - \tab 154 \tab 145 \tab 180 \tab 106 \tab 83 \tab 17 \tab 9 \tab 9 \tab 5 \tab 4 \tab 822 \tab - \tab 1534 \cr
   Total \tab 0 \tab 1604 \tab 1573 \tab 1991 \tab 1163 \tab 1049 \tab 300 \tab 151 \tab 145 \tab 82 \tab 138 \tab 1301 \tab 0 \tab 9497 \cr
}
}

\seealso{
Other War and International: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{egypt}},
\code{\link{israel}},
\code{\link{japan}}
}
\concept{War and International}
\keyword{variable}
