% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{engnrsci}
\alias{engnrsci}
\title{How scientific: engineering}
\source{
General Social Survey https://gss.norc.org
}
\description{
engnrsci
}
\details{
Question Please look at Card B20. How scientific are each of the following fields? If you have not heard of a particular field, just say you haven’t heard of it
H. Engineering. Is engineering very scientific, pretty scientific, not too scientific, or not scientific at all?
}
\section{Values}{

\itemize{
\item \code{1} very scientific
\item \code{2} pretty scientific
\item \code{3} not too scientific
\item \code{4} not scientific at all
\item \code{5} haven't heard of it (volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3413/vshow}{GSS Data Explorer page for \code{engnrsci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab haven't heard of it (volunteered) \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 846 \tab 604 \tab 210 \tab 125 \tab 8 \tab 67 \tab 4 \tab 4510 \cr
   2012 \tab 1477 \tab 245 \tab 160 \tab 50 \tab 20 \tab 2 \tab 16 \tab 4 \tab 1974 \cr
   Total \tab 4123 \tab 1091 \tab 764 \tab 260 \tab 145 \tab 10 \tab 83 \tab 8 \tab 6484 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{biosci}},
\code{\link{econsci}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{medsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
