% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_389.R
\name{engodd}
\alias{engodd}
\title{Engineers odd and peculiar}
\source{
General Social Survey https://gss.norc.org
}
\description{
engodd
}
\details{
Question F. Engineers are apt to be odd and peculiar people. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3531/vshow}{GSS Data Explorer page for \code{engodd}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{engint}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{farming}},
\code{\link{journlsm}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
