% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_254.R
\name{engoff1}
\alias{engoff1}
\title{English official language of u.s.}
\source{
General Social Survey https://gss.norc.org
}
\description{
engoff1
}
\details{
Question Do you favor a law making English the official language of the United States, or do you oppose such a law?
}
\section{Values}{

\itemize{
\item \verb{[1]} favor
\item \verb{[2]} oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2294/vshow}{GSS Data Explorer page for \code{engoff1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab favor \tab oppose \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 1013 \tab 294 \tab 82 \tab 9 \tab 2817 \cr
   Total \tab 1419 \tab 1013 \tab 294 \tab 82 \tab 9 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang1}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{twolang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
