% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_388.R
\name{engresp}
\alias{engresp}
\title{Consider work in engineer field}
\source{
General Social Survey https://gss.norc.org
}
\description{
engresp
}
\details{
Question C. Have you ever considered working in an engineering-related career?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3524/vshow}{GSS Data Explorer page for \code{engresp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab can't choose \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 126 \tab 355 \tab 1 \tab 3 \tab 1974 \cr
   Total \tab 1489 \tab 126 \tab 355 \tab 1 \tab 3 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
