% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_254.R
\name{engunite}
\alias{engunite}
\title{English unites all americans}
\source{
General Social Survey https://gss.norc.org
}
\description{
engunite
}
\details{
Question Now please tell us whether you strongly agree, agree, disagree or strongly disagree with each of these statements:
C. Speaking English as the common national language is what unites all Americans.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2297/vshow}{GSS Data Explorer page for \code{engunite}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engvote}},
\code{\link{forlang1}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{twolang}}

Other Language: 
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engvote}},
\code{\link{forlang1}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{twolang}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
