% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{enjoynat}
\alias{enjoynat}
\title{How much do you enjoy being outside in nature?}
\source{
General Social Survey https://gss.norc.org
}
\description{
enjoynat
}
\details{
Question How much, if at all, do you enjoy being outside in nature?
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[2]} to a small extent
\item \verb{[3]} to some extent
\item \verb{[4]} to a great extent
\item \verb{[5]} to a very great extent
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7495/vshow}{GSS Data Explorer page for \code{enjoynat}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab not at all \tab to a small extent \tab to some extent \tab to a great extent \tab to a very great extent \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 11 \tab 61 \tab 380 \tab 679 \tab 661 \tab 42 \tab 191 \tab 4 \tab 4032 \cr
   Total \tab 2003 \tab 11 \tab 61 \tab 380 \tab 679 \tab 661 \tab 42 \tab 191 \tab 4 \tab 4032 \cr
}
}

\keyword{variable}
