% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{enprbus}
\alias{enprbus}
\title{Most important environment problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
enprbus
}
\details{
Question D. Here is a list of some different environmental problems. Which problem, if any, do you think is the most important for America as a whole?
}
\section{Values}{

\itemize{
\item \code{1} air pollution
\item \code{2} chemicals and pesticides
\item \code{3} water shortage
\item \code{4} water pollution
\item \code{5} nuclear waste
\item \code{6} domestic waste disposal
\item \code{7} climate change
\item \code{8} genetically modified foods
\item \code{9} using up our natural resources
\item \code{10} none of these
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4803/vshow}{GSS Data Explorer page for \code{enprbus}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}

Other Environment: 
\code{\link{enprbfam}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
