% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_342.R
\name{entity}
\alias{entity}
\title{Entity mentions in religious change}
\source{
General Social Survey https://gss.norc.org
}
\description{
entity
}
\details{
Question Open-ended questions for how people who had undergone spiritual or religious change described their experiences. Specific religious figures mentioned in religious change.
}
\section{Values}{

\itemize{
\item \verb{[96]} no mention
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3067/vshow}{GSS Data Explorer page for \code{entity}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab no mention \tab no answer \tab Total \cr
   2004 \tab 2163 \tab 121 \tab 51 \tab 8 \tab 18 \tab 1 \tab 418 \tab 32 \tab 2812 \cr
   Total \tab 2163 \tab 121 \tab 51 \tab 8 \tab 18 \tab 1 \tab 418 \tab 32 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{agerborn}},
\code{\link{chngrbrn}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{numrborn}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}}
}
\concept{Spirituality and Religion}
\keyword{variable}
