% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{eqclose}
\alias{eqclose}
\title{R feels equally close to all persons}
\source{
General Social Survey https://gss.norc.org
}
\description{
eqclose
}
\details{
Question Do you feel equally close to all these people?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/849/vshow}{GSS Data Explorer page for \code{eqclose}}.
}

\seealso{
Other Social Networks: 
\code{\link{numgiven}}
}
\concept{Social Networks}
\keyword{variable}
