% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_28.R
\name{equal1}
\alias{equal1}
\title{Everyone should look out for self}
\source{
General Social Survey https://gss.norc.org
}
\description{
equal1
}
\details{
Question SHUFFLE THE GREY CARDS. READ STATEMENTS IN ORDER OF CARDS TO RESPONDENT. CIRCLE APPROPRIATE CODE FOR EACH STATEMENT.
A. In our society everyone must look out for himself. It is of little use to unite with others and fight for one's goals in politics or in unions.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} somewhat agree
\item \verb{[3]} somewhat disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/225/vshow}{GSS Data Explorer page for \code{equal1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 117 \tab 341 \tab 497 \tab 465 \tab 34 \tab 19 \tab 1473 \cr
   Total \tab 0 \tab 117 \tab 341 \tab 497 \tab 465 \tab 34 \tab 19 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{natmassz}},
\code{\link{natparkz}},
\code{\link{natroadz}},
\code{\link{natsocz}}
}
\concept{Core}
\keyword{variable}
