% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_29.R
\name{equal7}
\alias{equal7}
\title{Business profits are distributed fairly}
\source{
General Social Survey https://gss.norc.org
}
\description{
equal7
}
\details{
Question SHUFFLE THE GREY CARDS. READ STATEMENTS IN ORDER OF CARDS TO RESPONDENT. CIRCLE APPROPRIATE CODE FOR EACH STATEMENT.
G. Generally speaking, business profits are distributed fairly in the United States.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/231/vshow}{GSS Data Explorer page for \code{equal7}}.
}

\seealso{
Other Core: 
\code{\link{equal8}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}

Other Social Class: 
\code{\link{equal8}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}
}
\concept{Core}
\concept{Social Class}
\keyword{variable}
