% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_29.R
\name{equal8}
\alias{equal8}
\title{Social standing due to ability}
\source{
General Social Survey https://gss.norc.org
}
\description{
equal8
}
\details{
Question SHUFFLE THE GREY CARDS. READ STATEMENTS IN ORDER OF CARDS TO RESPONDENT. CIRCLE APPROPRIATE CODE FOR EACH STATEMENT.
H. If someone has a high social or economic position, that indicates the person has special abilities or great accomplishments.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/232/vshow}{GSS Data Explorer page for \code{equal8}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 172 \tab 531 \tab 450 \tab 254 \tab 48 \tab 18 \tab 1473 \cr
   Total \tab 0 \tab 172 \tab 531 \tab 450 \tab 254 \tab 48 \tab 18 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{equal7}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}
}
\concept{Core}
\keyword{variable}
