% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{eqwlthy}
\alias{eqwlthy}
\title{Govt reduce income differences-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
eqwlthy
}
\details{
Question B. Some people think that the income differences between the rich and the poor ought to be reduced, perhaps by raisingthe taxes of wealthy families or by giving income assistance to the poor. Others think that the government should not concern itself with reducing this income difference between the rich and the poor. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that the government ought to reduce the income differences between rich and poor, and a score of 7 meaning that the government should not concern itself with reducing income differences. What score between 1 and 7 comes closest to the way you feel? (CIRCLE ONE):
}
\section{Values}{

\itemize{
\item \code{1} govt reduce diff
\item \code{7} no govt action
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/244/vshow}{GSS Data Explorer page for \code{eqwlthy}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab govt reduce diff \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no govt action \tab don't know \tab no answer \tab Total \cr
   1978 \tab 761 \tab 145 \tab 62 \tab 144 \tab 170 \tab 84 \tab 60 \tab 84 \tab 19 \tab 3 \tab 1532 \cr
   Total \tab 761 \tab 145 \tab 62 \tab 144 \tab 170 \tab 84 \tab 60 \tab 84 \tab 19 \tab 3 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}},
\code{\link{tax}}
}
\concept{Core}
\keyword{variable}
