% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{erafirm}
\alias{erafirm}
\title{How firm is r's opinion on era}
\source{
General Social Survey https://gss.norc.org
}
\description{
erafirm
}
\details{
Question How firm are you about (your opinion on/the way you feel about) the Equal Rights Amendment--would you say you are very likely to change (your opinion), somewhat likely to change, somewhat unlikely to change or very unlikely to change your opinion?
}
\section{Values}{

\itemize{
\item \verb{[1]} change very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} change very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/716/vshow}{GSS Data Explorer page for \code{erafirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab change very likely \tab somewhat likely \tab somewhat unlikely \tab change very unlikely \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 66 \tab 356 \tab 525 \tab 824 \tab 80 \tab 9 \tab 1860 \cr
   Total \tab 0 \tab 66 \tab 356 \tab 525 \tab 824 \tab 80 \tab 9 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubgv}},
\code{\link{racsubs}}
}
\concept{Core}
\keyword{variable}
