% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{eraimp}
\alias{eraimp}
\title{Importance of era issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
eraimp
}
\details{
Question How important is the Equal Rights Amendment issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most imp
\item \verb{[2]} important
\item \verb{[3]} not very imp
\item \verb{[4]} not important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/714/vshow}{GSS Data Explorer page for \code{eraimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not important \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 223 \tab 815 \tab 565 \tab 213 \tab 39 \tab 5 \tab 1860 \cr
   Total \tab 0 \tab 223 \tab 815 \tab 565 \tab 213 \tab 39 \tab 5 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubgv}},
\code{\link{racsubs}}
}
\concept{Core}
\keyword{variable}
