% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{erainfo}
\alias{erainfo}
\title{How much info does r have about era}
\source{
General Social Survey https://gss.norc.org
}
\description{
erainfo
}
\details{
Question How much information do you have about the Equal Rights Amendment? Do you have all the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \verb{[1]} all info r needs
\item \verb{[2]} most of the info
\item \verb{[3]} some info
\item \verb{[4]} very little info
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/715/vshow}{GSS Data Explorer page for \code{erainfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all info r needs \tab most of the info \tab some info \tab very little info \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 216 \tab 305 \tab 761 \tab 555 \tab 17 \tab 6 \tab 1860 \cr
   Total \tab 0 \tab 216 \tab 305 \tab 761 \tab 555 \tab 17 \tab 6 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubgv}},
\code{\link{racsubs}}
}
\concept{Core}
\keyword{variable}
