% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{erameans}
\alias{erameans}
\title{Does r understand what era means}
\source{
General Social Survey https://gss.norc.org
}
\description{
erameans
}
\details{
Question Have you heard or read about the Equal Rights Amendment?
A. Do you understand what the Equal Rights Amendment means?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/708/vshow}{GSS Data Explorer page for \code{erameans}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab 209 \tab 1060 \tab 239 \tab 22 \tab - \tab 1530 \cr
   1982 \tab 183 \tab 1405 \tab 230 \tab 35 \tab 7 \tab 1860 \cr
   Total \tab 392 \tab 2465 \tab 469 \tab 57 \tab 7 \tab 3390 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{ablegal}},
\code{\link{era}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
