% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{eraread}
\alias{eraread}
\title{Has r heard or read about era}
\source{
General Social Survey https://gss.norc.org
}
\description{
eraread
}
\details{
Question Have you heard or read about the Equal Rights Amendment?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/707/vshow}{GSS Data Explorer page for \code{eraread}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 1321 \tab 188 \tab 18 \tab 3 \tab 1530 \cr
   1982 \tab - \tab 1676 \tab 138 \tab 45 \tab 1 \tab 1860 \cr
   Total \tab 0 \tab 2997 \tab 326 \tab 63 \tab 4 \tab 3390 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{ablegal}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
