% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{eratell}
\alias{eratell}
\title{Does r favor or oppose era}
\source{
General Social Survey https://gss.norc.org
}
\description{
eratell
}
\details{
Question Have you heard or read about the Equal Rights Amendment?
C. The Equal Rights Amendment will amend the United States Constitution to provide equal rights for men and women. Do you strongly favor, somewhat favor, somewhat oppose, or strongly oppose this amendment?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} somewhat favor
\item \verb{[3]} somewht oppose
\item \verb{[4]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/710/vshow}{GSS Data Explorer page for \code{eratell}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly favor \tab somewhat favor \tab somewht oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1676 \tab 47 \tab 73 \tab 21 \tab 11 \tab 31 \tab 1 \tab 1860 \cr
   Total \tab 1676 \tab 47 \tab 73 \tab 21 \tab 11 \tab 31 \tab 1 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{ablegal}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
