% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{esop}
\alias{esop}
\title{R is member of esop}
\source{
General Social Survey https://gss.norc.org
}
\description{
esop
}
\details{
Question A. Are you a member of an Employee Stock Ownership Plan commonly called an ESOP?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018 \cr
   B/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3561/vshow}{GSS Data Explorer page for \code{esop}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2014 \tab 2432 \tab 22 \tab 81 \tab 1 \tab 2 \tab - \tab 2538 \cr
   2018 \tab 2142 \tab 58 \tab 136 \tab 10 \tab 2 \tab - \tab 2348 \cr
   2022 \tab 3423 \tab 25 \tab 83 \tab 11 \tab - \tab 2 \tab 3544 \cr
   Total \tab 7997 \tab 105 \tab 300 \tab 22 \tab 4 \tab 2 \tab 8430 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
