% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{esopnot}
\alias{esopnot}
\title{Would r rather work for a company with an esop or cash sharing}
\source{
General Social Survey https://gss.norc.org
}
\description{
esopnot
}
\details{
Question If you are given two job offers for the same position with the same salary and benefits, one in a company that also has an ESOP (Employee Stock Ownership Plan) or cash profit sharing for its employees and the other in a company that does not have an ESOP (Employee Stock Ownership Plan) or cash profit sharing for its employees, which job offer are you most inclined to take?
}
\section{Values}{

\itemize{
\item \code{1} i would definitely take the job without the esop (employee stock ownership plan) or cash profit sharing
\item \code{2} i would probably take the job without the esop (employee stock ownership plan) or cash profit sharing
\item \code{3} i would be neutral
\item \code{4} i would probably take the job with the esop (employee stock ownership plan) or cash profit sharing
\item \code{5} i would definitely take the job with the esop (employee stock ownership plan) or cash profit sharing
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7226/vshow}{GSS Data Explorer page for \code{esopnot}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab i would definitely take the job without the esop (employee stock ownership plan) or cash profit sharing \tab i would probably take the job without the esop (employee stock ownership plan) or cash profit sharing \tab i would be neutral \tab i would probably take the job with the esop (employee stock ownership plan) or cash profit sharing \tab i would definitely take the job with the esop (employee stock ownership plan) or cash profit sharing \tab don't know \tab no answer \tab Total \cr
   2018 \tab 900 \tab 97 \tab 45 \tab 384 \tab 437 \tab 422 \tab 28 \tab 35 \tab 2348 \cr
   Total \tab 900 \tab 97 \tab 45 \tab 384 \tab 437 \tab 422 \tab 28 \tab 35 \tab 2348 \cr
}
}

\keyword{variable}
