% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{eth1}
\alias{eth1}
\title{1st mentioned country of origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
eth1
}
\details{
Question From what countries or part of the world did your ancestors come?
B. FIRST ETHNICITY MENTIONED
}
\section{Values}{

\itemize{
\item \verb{[1]} africa (general)
\item \verb{[2]} austria
\item \verb{[3]} french canada
\item \verb{[4]} other canada
\item \verb{[5]} china
\item \verb{[6]} czechoslovakia
\item \verb{[7]} denmark
\item \verb{[8]} england & wales
\item \verb{[9]} finland
\item \verb{[10]} france
\item \verb{[11]} germany
\item \verb{[12]} greece
\item \verb{[13]} hungary
\item \verb{[14]} ireland
\item \verb{[15]} italy
\item \verb{[16]} japan
\item \verb{[17]} mexico
\item \verb{[18]} netherlands
\item \verb{[19]} norway
\item \verb{[20]} philippines
\item \verb{[21]} poland
\item \verb{[22]} puerto rico
\item \verb{[23]} russia
\item \verb{[24]} scotland
\item \verb{[25]} spain
\item \verb{[26]} sweden
\item \verb{[27]} switzerland
\item \verb{[28]} west indies
\item \verb{[29]} other
\item \verb{[30]} american indian
\item \verb{[31]} india
\item \verb{[32]} portugal
\item \verb{[33]} lithuania
\item \verb{[34]} yugoslavia
\item \verb{[35]} romania
\item \verb{[36]} belgium
\item \verb{[37]} arabic
\item \verb{[38]} other spanish
\item \verb{[39]} non-span west indies
\item \verb{[40]} other asian
\item \verb{[41]} other european
\item \verb{[97]} american only
\item \verb{[101]} turkey
\item \verb{[202]} algeria
\item \verb{[203]} congo
\item \verb{[204]} egypt
\item \verb{[205]} ethiopia
\item \verb{[206]} kenya
\item \verb{[207]} nigeria
\item \verb{[208]} south africa
\item \verb{[209]} sudan
\item \verb{[210]} tanzania
\item \verb{[211]} uganda
\item \verb{[299]} other africa
\item \verb{[301]} south korea
\item \verb{[302]} bangladesh
\item \verb{[303]} indonesia
\item \verb{[304]} pakistan
\item \verb{[306]} thailand
\item \verb{[307]} vietnam
\item \verb{[401]} iran
\item \verb{[402]} iraq
\item \verb{[403]} israel
\item \verb{[404]} jordan
\item \verb{[405]} saudi arabia
\item \verb{[406]} syria
\item \verb{[407]} uae
\item \verb{[408]} yemen
\item \verb{[499]} other middle east
\item \verb{[501]} argentina
\item \verb{[502]} bolivia
\item \verb{[503]} brazil
\item \verb{[504]} chile
\item \verb{[505]} colombia
\item \verb{[506]} ecuador
\item \verb{[507]} french guiana
\item \verb{[508]} guyana
\item \verb{[509]} paraguay
\item \verb{[510]} peru
\item \verb{[511]} suriname
\item \verb{[512]} uruguay
\item \verb{[513]} venezuela
\item \verb{[599]} other south america
\item \verb{[601]} belize
\item \verb{[602]} costa rica
\item \verb{[603]} el salvador
\item \verb{[604]} guatemala
\item \verb{[605]} honduras
\item \verb{[606]} nicaragua
\item \verb{[607]} panama
\item \verb{[699]} other central america
\item \verb{[799]} other north america
\item \verb{[801]} cuba
\item \verb{[802]} haiti
\item \verb{[803]} dominican republic
\item \verb{[804]} jamaica
\item \verb{[899]} other caribbean
\item \verb{[901]} australia
\item \verb{[902]} papua new guinea
\item \verb{[903]} new zealand
\item \verb{[904]} samoa
\item \verb{[999]} other oceania
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5264/vshow}{GSS Data Explorer page for \code{eth1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa (general) \tab austria \tab french canada \tab other canada \tab china \tab czechoslovakia \tab denmark \tab england & wales \tab finland \tab france \tab germany \tab greece \tab hungary \tab ireland \tab italy \tab japan \tab mexico \tab netherlands \tab norway \tab philippines \tab poland \tab puerto rico \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab west indies \tab other \tab american indian \tab india \tab portugal \tab lithuania \tab yugoslavia \tab romania \tab belgium \tab other spanish \tab non-span west indies \tab american only \tab don't know \tab no answer \tab arabic \tab other asian \tab other european \tab iap \tab congo \tab egypt \tab ethiopia \tab kenya \tab nigeria \tab south africa \tab sudan \tab uganda \tab other africa \tab south korea \tab pakistan \tab thailand \tab other middle east \tab argentina \tab brazil \tab colombia \tab belize \tab el salvador \tab panama \tab other north america \tab cuba \tab haiti \tab dominican republic \tab jamaica \tab samoa \tab other oceania \tab skipped on web \tab Total \cr
   1984 \tab - \tab 90 \tab 8 \tab 24 \tab 7 \tab 6 \tab 13 \tab 13 \tab 199 \tab 3 \tab 36 \tab 250 \tab 7 \tab 14 \tab 165 \tab 68 \tab 6 \tab 28 \tab 41 \tab 28 \tab 4 \tab 50 \tab 13 \tab 22 \tab 53 \tab 16 \tab 17 \tab 10 \tab 1 \tab 11 \tab 47 \tab 3 \tab 3 \tab 3 \tab 7 \tab 2 \tab 3 \tab 7 \tab 6 \tab 14 \tab 170 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 67 \tab 10 \tab 27 \tab 6 \tab 5 \tab 23 \tab 10 \tab 238 \tab 5 \tab 55 \tab 277 \tab 4 \tab 6 \tab 148 \tab 86 \tab 6 \tab 36 \tab 36 \tab 34 \tab 2 \tab 40 \tab 9 \tab 18 \tab 51 \tab 20 \tab 35 \tab 5 \tab 1 \tab 18 \tab 41 \tab 1 \tab 2 \tab 2 \tab 6 \tab 5 \tab 4 \tab 9 \tab 1 \tab 15 \tab 162 \tab 4 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 95 \tab 14 \tab 16 \tab 11 \tab 4 \tab 18 \tab 12 \tab 200 \tab 3 \tab 40 \tab 252 \tab 6 \tab 6 \tab 171 \tab 82 \tab 5 \tab 31 \tab 29 \tab 36 \tab 2 \tab 59 \tab 12 \tab 29 \tab 32 \tab 19 \tab 24 \tab 8 \tab 2 \tab 14 \tab 31 \tab 5 \tab 3 \tab 3 \tab 2 \tab 3 \tab - \tab 4 \tab 7 \tab 17 \tab 160 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 284 \tab 10 \tab 22 \tab 9 \tab 8 \tab 10 \tab 9 \tab 190 \tab 5 \tab 39 \tab 289 \tab 4 \tab 5 \tab 147 \tab 65 \tab 3 \tab 34 \tab 22 \tab 22 \tab 2 \tab 42 \tab 5 \tab 21 \tab 62 \tab 17 \tab 28 \tab 7 \tab 3 \tab 18 \tab 85 \tab 6 \tab 9 \tab 5 \tab 9 \tab 2 \tab 2 \tab 7 \tab 8 \tab 36 \tab 254 \tab 10 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 89 \tab 9 \tab 16 \tab 13 \tab 6 \tab 17 \tab 8 \tab 188 \tab 1 \tab 36 \tab 284 \tab 8 \tab 9 \tab 167 \tab 53 \tab 4 \tab 42 \tab 23 \tab 25 \tab 1 \tab 36 \tab 15 \tab 23 \tab 52 \tab 19 \tab 28 \tab 6 \tab - \tab 18 \tab 46 \tab 7 \tab 1 \tab 5 \tab 7 \tab 2 \tab 1 \tab 7 \tab 2 \tab 26 \tab 169 \tab 9 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 87 \tab 9 \tab 17 \tab 9 \tab 9 \tab 18 \tab 12 \tab 230 \tab 2 \tab 39 \tab 330 \tab 5 \tab 8 \tab 153 \tab 77 \tab 6 \tab 37 \tab 29 \tab 38 \tab 5 \tab 41 \tab 9 \tab 29 \tab 60 \tab 11 \tab 11 \tab 7 \tab 2 \tab 16 \tab 48 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 11 \tab 5 \tab 13 \tab 125 \tab 13 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 107 \tab 7 \tab 23 \tab 9 \tab 4 \tab 21 \tab 13 \tab 185 \tab 5 \tab 24 \tab 240 \tab 10 \tab 4 \tab 139 \tab 53 \tab 2 \tab 32 \tab 20 \tab 22 \tab 1 \tab 41 \tab 12 \tab 32 \tab 52 \tab 10 \tab 25 \tab 8 \tab - \tab 13 \tab 51 \tab 3 \tab 1 \tab 9 \tab 6 \tab 4 \tab 4 \tab 9 \tab 8 \tab 6 \tab 146 \tab 8 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 117 \tab 13 \tab 26 \tab 10 \tab 3 \tab 18 \tab 13 \tab 214 \tab 2 \tab 49 \tab 240 \tab 5 \tab 9 \tab 171 \tab 74 \tab 5 \tab 43 \tab 20 \tab 27 \tab 4 \tab 43 \tab 13 \tab 27 \tab 47 \tab 15 \tab 23 \tab 13 \tab - \tab 20 \tab 43 \tab 2 \tab 3 \tab 3 \tab 9 \tab 2 \tab 3 \tab 6 \tab 7 \tab 16 \tab 152 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 105 \tab 8 \tab 15 \tab 8 \tab 7 \tab 14 \tab 11 \tab 189 \tab 5 \tab 40 \tab 304 \tab 8 \tab 17 \tab 166 \tab 65 \tab 3 \tab 28 \tab 39 \tab 19 \tab 8 \tab 47 \tab 11 \tab 37 \tab 73 \tab 24 \tab 25 \tab 7 \tab 1 \tab 18 \tab 58 \tab 7 \tab 5 \tab 5 \tab 3 \tab 1 \tab 1 \tab 10 \tab 11 \tab 22 \tab 175 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 231 \tab 23 \tab 42 \tab 16 \tab 9 \tab 47 \tab 21 \tab 368 \tab 5 \tab 71 \tab 602 \tab 13 \tab 13 \tab 321 \tab 122 \tab 6 \tab 68 \tab 50 \tab 49 \tab 12 \tab 72 \tab 20 \tab 41 \tab 123 \tab 32 \tab 37 \tab 12 \tab 4 \tab 12 \tab 81 \tab 14 \tab 8 \tab 12 \tab 5 \tab 7 \tab 3 \tab 20 \tab 16 \tab 74 \tab 272 \tab 7 \tab 8 \tab 17 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 195 \tab 21 \tab 27 \tab 10 \tab 12 \tab 38 \tab 13 \tab 348 \tab 8 \tab 76 \tab 527 \tab 15 \tab 18 \tab 264 \tab 133 \tab 4 \tab 72 \tab 42 \tab 38 \tab 11 \tab 62 \tab 25 \tab 43 \tab 105 \tab 35 \tab 43 \tab 11 \tab - \tab 26 \tab 94 \tab 13 \tab 10 \tab 6 \tab 13 \tab 5 \tab 3 \tab 24 \tab 12 \tab 65 \tab 354 \tab 53 \tab 4 \tab 17 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 216 \tab 12 \tab 45 \tab 13 \tab 24 \tab 33 \tab 17 \tab 321 \tab 5 \tab 83 \tab 529 \tab 11 \tab 24 \tab 291 \tab 135 \tab 6 \tab 75 \tab 42 \tab 49 \tab 20 \tab 82 \tab 33 \tab 24 \tab 94 \tab 26 \tab 34 \tab 8 \tab 1 \tab 27 \tab 82 \tab 9 \tab 9 \tab 9 \tab 14 \tab 3 \tab 6 \tab 31 \tab 19 \tab 67 \tab 259 \tab 16 \tab 3 \tab 14 \tab 11 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 234 \tab 18 \tab 40 \tab 13 \tab 8 \tab 27 \tab 18 \tab 320 \tab 8 \tab 60 \tab 456 \tab 19 \tab 8 \tab 297 \tab 132 \tab 8 \tab 92 \tab 36 \tab 55 \tab 27 \tab 79 \tab 37 \tab 53 \tab 94 \tab 21 \tab 40 \tab 10 \tab 6 \tab 31 \tab 127 \tab 17 \tab 8 \tab 9 \tab 9 \tab 3 \tab 4 \tab 30 \tab 14 \tab 55 \tab 230 \tab 37 \tab 2 \tab 17 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 241 \tab 24 \tab 34 \tab 13 \tab 12 \tab 17 \tab 19 \tab 297 \tab 7 \tab 76 \tab 523 \tab 17 \tab 22 \tab 282 \tab 127 \tab 7 \tab 81 \tab 34 \tab 44 \tab 17 \tab 86 \tab 34 \tab 53 \tab 96 \tab 34 \tab 51 \tab 7 \tab 4 \tab 5 \tab 77 \tab 14 \tab 7 \tab 10 \tab 5 \tab 8 \tab 10 \tab 28 \tab 16 \tab 59 \tab 219 \tab 7 \tab 7 \tab 17 \tab 17 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 233 \tab 10 \tab 12 \tab 13 \tab 27 \tab 28 \tab 19 \tab 348 \tab 12 \tab 58 \tab 440 \tab 11 \tab 12 \tab 338 \tab 130 \tab 11 \tab 132 \tab 30 \tab 39 \tab 24 \tab 54 \tab 27 \tab 47 \tab 88 \tab 28 \tab 35 \tab 13 \tab 1 \tab 10 \tab 134 \tab 23 \tab 12 \tab 6 \tab 8 \tab 6 \tab 7 \tab 29 \tab 14 \tab 57 \tab 219 \tab 12 \tab 13 \tab 24 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 407 \tab 14 \tab 18 \tab 18 \tab 28 \tab 57 \tab 33 \tab 457 \tab 13 \tab 102 \tab 737 \tab 19 \tab 27 \tab 475 \tab 215 \tab 19 \tab 327 \tab 55 \tab 55 \tab 26 \tab 112 \tab 49 \tab 62 \tab 159 \tab 71 \tab 73 \tab 22 \tab 5 \tab 20 \tab 104 \tab 29 \tab 19 \tab 11 \tab 14 \tab 6 \tab 13 \tab 99 \tab 26 \tab 125 \tab 314 \tab 14 \tab 13 \tab 28 \tab 20 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 156 \tab 10 \tab 8 \tab 11 \tab 14 \tab 18 \tab 11 \tab 226 \tab 5 \tab 45 \tab 338 \tab 5 \tab 11 \tab 221 \tab 97 \tab 5 \tab 126 \tab 22 \tab 29 \tab 14 \tab 43 \tab 19 \tab 26 \tab 54 \tab 36 \tab 33 \tab 11 \tab 3 \tab 9 \tab 53 \tab 24 \tab 13 \tab 10 \tab 5 \tab 5 \tab 7 \tab 42 \tab 21 \tab 42 \tab 160 \tab 8 \tab 3 \tab 15 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 164 \tab 2 \tab 15 \tab 4 \tab 12 \tab 17 \tab 17 \tab 212 \tab 6 \tab 43 \tab 359 \tab 9 \tab 6 \tab 223 \tab 87 \tab 12 \tab 124 \tab 26 \tab 24 \tab 10 \tab 57 \tab 21 \tab 30 \tab 60 \tab 28 \tab 16 \tab 5 \tab 2 \tab 8 \tab 53 \tab 13 \tab 19 \tab 5 \tab 5 \tab 5 \tab 7 \tab 39 \tab 15 \tab 56 \tab 185 \tab 6 \tab 5 \tab 21 \tab 11 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 178 \tab 10 \tab 15 \tab 8 \tab 16 \tab 10 \tab 10 \tab 187 \tab 10 \tab 41 \tab 327 \tab 14 \tab 8 \tab 204 \tab 93 \tab 2 \tab 131 \tab 16 \tab 22 \tab 12 \tab 39 \tab 17 \tab 23 \tab 48 \tab 51 \tab 24 \tab 9 \tab 2 \tab 19 \tab 64 \tab 24 \tab 12 \tab 5 \tab 8 \tab 4 \tab 2 \tab 38 \tab 13 \tab 63 \tab 158 \tab 5 \tab 4 \tab 19 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 201 \tab 14 \tab 23 \tab 8 \tab 23 \tab 18 \tab 9 \tab 256 \tab 10 \tab 46 \tab 418 \tab 14 \tab 14 \tab 267 \tab 132 \tab 9 \tab 191 \tab 25 \tab 24 \tab 17 \tab 59 \tab 30 \tab 29 \tab 66 \tab 47 \tab 38 \tab 4 \tab 3 \tab 5 \tab 67 \tab 31 \tab 12 \tab 9 \tab 4 \tab 7 \tab 3 \tab 76 \tab 23 \tab 47 \tab 209 \tab 8 \tab 8 \tab 15 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 268 \tab 6 \tab 21 \tab 15 \tab 23 \tab 32 \tab 14 \tab 244 \tab 9 \tab 59 \tab 455 \tab 17 \tab 14 \tab 274 \tab 149 \tab 8 \tab 184 \tab 24 \tab 35 \tab 23 \tab 84 \tab 35 \tab 42 \tab 86 \tab 47 \tab 41 \tab 13 \tab 3 \tab 12 \tab 90 \tab 26 \tab 14 \tab 13 \tab 12 \tab 5 \tab 8 \tab 65 \tab 15 \tab 57 \tab 271 \tab 6 \tab 12 \tab 21 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 196 \tab 14 \tab 18 \tab 13 \tab 19 \tab 22 \tab 8 \tab 192 \tab 12 \tab 53 \tab 360 \tab 10 \tab 9 \tab 210 \tab 97 \tab 6 \tab 180 \tab 13 \tab 28 \tab 18 \tab 55 \tab 26 \tab 35 \tab 63 \tab 46 \tab 24 \tab 10 \tab 1 \tab 11 \tab 62 \tab 20 \tab 7 \tab 8 \tab 4 \tab 4 \tab 3 \tab 51 \tab 20 \tab 61 \tab 292 \tab 8 \tab 8 \tab 23 \tab 28 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 2 \tab 4 \tab 6 \tab 1 \tab 10 \tab 7 \tab 144 \tab 3 \tab 37 \tab 136 \tab 4 \tab 2 \tab 136 \tab 31 \tab 2 \tab 6 \tab 11 \tab 18 \tab - \tab 25 \tab 4 \tab 11 \tab 66 \tab 25 \tab 23 \tab 5 \tab - \tab - \tab 29 \tab 1 \tab 8 \tab 9 \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab 17 \tab 2 \tab - \tab 1 \tab 11 \tab 2683 \tab 1 \tab 5 \tab 1 \tab 2 \tab 7 \tab 3 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3544 \cr
   Total \tab 0 \tab 3961 \tab 268 \tab 508 \tab 243 \tab 280 \tab 526 \tab 317 \tab 5753 \tab 144 \tab 1208 \tab 8673 \tab 235 \tab 266 \tab 5230 \tab 2303 \tab 145 \tab 2100 \tab 685 \tab 760 \tab 260 \tab 1308 \tab 476 \tab 757 \tab 1684 \tab 678 \tab 728 \tab 211 \tab 45 \tab 341 \tab 1567 \tab 296 \tab 189 \tab 159 \tab 158 \tab 91 \tab 101 \tab 642 \tab 279 \tab 993 \tab 4672 \tab 249 \tab 110 \tab 249 \tab 191 \tab 6715 \tab 1 \tab 5 \tab 1 \tab 2 \tab 7 \tab 3 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 56811 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}}
}
\concept{Ethnicity}
\keyword{variable}
