% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_257.R
\name{ethchng}
\alias{ethchng}
\title{Ethnic change in 25 yrs good for country}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethchng
}
\details{
Question When you think about these changes in the racial and ethnic make-up of the country in the next 25 years, do you think they will be a very good thing for the country, a good thing, neither good nor bad, a bad thing, or a very bad thing?
}
\section{Values}{

\itemize{
\item \verb{[1]} very good thing
\item \verb{[2]} good thing
\item \verb{[3]} neither good nor bad
\item \verb{[4]} bad thing
\item \verb{[5]} very bad thing
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2322/vshow}{GSS Data Explorer page for \code{ethchng}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very good thing \tab good thing \tab neither good nor bad \tab bad thing \tab very bad thing \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 42 \tab 301 \tab 730 \tab 195 \tab 52 \tab 61 \tab 17 \tab 2817 \cr
   Total \tab 1419 \tab 42 \tab 301 \tab 730 \tab 195 \tab 52 \tab 61 \tab 17 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{hispwork}},
\code{\link{knwwht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
