% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{ethclose}
\alias{ethclose}
\title{How close feel to your ethnic or racial group?}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethclose
}
\details{
Question How close do you feel to your ethnic or racial group?
}
\section{Values}{

\itemize{
\item \verb{[1]} very close
\item \verb{[2]} close
\item \verb{[3]} not very close
\item \verb{[4]} not close at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4875/vshow}{GSS Data Explorer page for \code{ethclose}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very close \tab close \tab not very close \tab not close at all \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1537 \tab 490 \tab 474 \tab 216 \tab 67 \tab 96 \tab 24 \tab - \tab 2904 \cr
   2004 \tab 1596 \tab 467 \tab 410 \tab 205 \tab 50 \tab 79 \tab 5 \tab - \tab 2812 \cr
   Total \tab 3133 \tab 957 \tab 884 \tab 421 \tab 117 \tab 175 \tab 29 \tab 0 \tab 5716 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amproud1}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{belikeus}},
\code{\link{biasnews}},
\code{\link{bizmoney}},
\code{\link{brneffrt}},
\code{\link{citizen}},
\code{\link{citworld}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{decsorgs}},
\code{\link{experts}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{freetrde}},
\code{\link{fulldem}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immassim}},
\code{\link{immcosts}},
\code{\link{immcrime1}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{immideas}},
\code{\link{immideas1}},
\code{\link{immimp}},
\code{\link{immjobs1}},
\code{\link{immrghts}},
\code{\link{immwlfare}},
\code{\link{imports}},
\code{\link{incgap1}},
\code{\link{internet}},
\code{\link{intlincs}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{lessprd}},
\code{\link{letin1}},
\code{\link{letin1a}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{refugees}},
\code{\link{res161}},
\code{\link{shortcom}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{wrldgovt}}

Other Ethnicity: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethadapt}},
\code{\link{ethdiff}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethnic}},
\code{\link{ethnofit}},
\code{\link{ethnum}},
\code{\link{ethorgs}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{fucitzn}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hspgovt}},
\code{\link{meltpot}},
\code{\link{natborn}},
\code{\link{owneth}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{schleth}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{teacheth}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{whoteach}},
\code{\link{whtgovt}}
}
\concept{Ethnicity}
\concept{ISSP National Identity}
\keyword{variable}
