% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{ethhist}
\alias{ethhist}
\title{Minority history adequately taught in school}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethhist
}
\details{
Question In history classes in high school and college, do the experiences of racial and ethnic minority groups in America receive too much attention now, too little attention, or about the right amount?
}
\section{Values}{

\itemize{
\item \verb{[1]} too much attention now
\item \verb{[2]} too little attention
\item \verb{[3]} about the right amount
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1630/vshow}{GSS Data Explorer page for \code{ethhist}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too much attention now \tab too little attention \tab about the right amount \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 334 \tab 306 \tab 630 \tab 193 \tab 11 \tab 2992 \cr
   Total \tab 1518 \tab 334 \tab 306 \tab 630 \tab 193 \tab 11 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
