% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_170.R
\name{ethid}
\alias{ethid}
\title{R american vs ethnic identity}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethid
}
\details{
Question When you think of social and political issues, do you think of yourself mainly as a member of a particular ethnic, racial, or nationality group or do you think of yourself mainly as just an American?
}
\section{Values}{

\itemize{
\item \code{1} just an american
\item \code{2} some part. ethnic grp
\item \code{3} 2 or more ethnic grps
\item \code{4} amer and ethnic grps
\item \code{5} it depends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1591/vshow}{GSS Data Explorer page for \code{ethid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab just an american \tab some part. ethnic grp \tab 2 or more ethnic grps \tab amer and ethnic grps \tab it depends \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 1315 \tab 109 \tab 3 \tab 20 \tab 12 \tab 8 \tab 7 \tab 2992 \cr
   Total \tab 1518 \tab 1315 \tab 109 \tab 3 \tab 20 \tab 12 \tab 8 \tab 7 \tab 2992 \cr
}
}

\concept{Multiculturalism}
\keyword{variable}
