% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{ethimp}
\alias{ethimp}
\title{Importance of ethnic group to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethimp
}
\details{
Question When you think about yourself, how important is your ethnic group membership to your sense of who you are?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} moderately important
\item \verb{[3]} slightly important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2725/vshow}{GSS Data Explorer page for \code{ethimp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab moderately important \tab slightly important \tab not at all important \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2002 \tab - \tab 903 \tab 697 \tab 471 \tab 667 \tab 20 \tab 7 \tab - \tab 2765 \cr
   Total \tab 0 \tab 903 \tab 697 \tab 471 \tab 667 \tab 20 \tab 7 \tab 0 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethdiff}},
\code{\link{ethignor}},
\code{\link{ethnofit}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelblks}},
\code{\link{feelhsps}},
\code{\link{feelwhts}},
\code{\link{whtsdiff}}

Other Ethnicity: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethadapt}},
\code{\link{ethclose}},
\code{\link{ethdiff}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethignor}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethnic}},
\code{\link{ethnofit}},
\code{\link{ethnum}},
\code{\link{ethorgs}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{fucitzn}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hspgovt}},
\code{\link{meltpot}},
\code{\link{natborn}},
\code{\link{owneth}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{schleth}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{teacheth}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{whoteach}},
\code{\link{whtgovt}}
}
\concept{Ethnicity}
\concept{Prejudice}
\keyword{variable}
