% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{ethissu1}
\alias{ethissu1}
\title{Degree r invokes primary identity as ethnic}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethissu1
}
\details{
Question E. Do you think of yourself as a(n) (ethnic group or groups specified) on all issues, most issues, some issues, or just a few issues?
}
\section{Values}{

\itemize{
\item \verb{[1]} all issues
\item \verb{[2]} most issues
\item \verb{[3]} some issues
\item \verb{[4]} just a few issues
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1596/vshow}{GSS Data Explorer page for \code{ethissu1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab most issues \tab some issues \tab just a few issues \tab no answer \tab Total \cr
   1994 \tab 2972 \tab 8 \tab 2 \tab 1 \tab 9 \tab 2992 \cr
   Total \tab 2972 \tab 8 \tab 2 \tab 1 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{ethid1}},
\code{\link{ethissue}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{meltpot}}
}
\concept{Multiculturalism}
\keyword{variable}
