% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{ethnic}
\alias{ethnic}
\title{Country of family origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethnic
}
\details{
Question From what countries or part of the world did your ancestors come?
A. IF MORE THAN ONE COUNTRY NAMED:Which one of these countries do you feel closer to?
}
\section{Values}{

\itemize{
\item \code{1} africa (general)
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} romania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span west indies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{101} turkey
\item \code{202} algeria
\item \code{203} congo
\item \code{204} egypt
\item \code{205} ethiopia
\item \code{206} kenya
\item \code{207} nigeria
\item \code{208} south africa
\item \code{209} sudan
\item \code{210} tanzania
\item \code{211} uganda
\item \code{299} other africa
\item \code{301} south korea
\item \code{302} bangladesh
\item \code{303} indonesia
\item \code{304} pakistan
\item \code{306} thailand
\item \code{307} vietnam
\item \code{401} iran
\item \code{402} iraq
\item \code{403} israel
\item \code{404} jordan
\item \code{405} saudi arabia
\item \code{406} syria
\item \code{407} uae
\item \code{408} yemen
\item \code{499} other middle east
\item \code{501} argentina
\item \code{502} bolivia
\item \code{503} brazil
\item \code{504} chile
\item \code{505} colombia
\item \code{506} ecuador
\item \code{507} french guiana
\item \code{508} guyana
\item \code{509} paraguay
\item \code{510} peru
\item \code{511} suriname
\item \code{512} uruguay
\item \code{513} venezuela
\item \code{599} other south america
\item \code{601} belize
\item \code{602} costa rica
\item \code{603} el salvador
\item \code{604} guatemala
\item \code{605} honduras
\item \code{606} nicaragua
\item \code{607} panama
\item \code{699} other central america
\item \code{799} other north america
\item \code{801} cuba
\item \code{802} haiti
\item \code{803} dominican republic
\item \code{804} jamaica
\item \code{899} other caribbean
\item \code{901} australia
\item \code{902} papua new guinea
\item \code{903} new zealand
\item \code{904} samoa
\item \code{999} other oceania
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnum}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}}
}
\concept{Ethnicity}
\keyword{variable}
