% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_615.R
\name{ethregion21}
\alias{ethregion21}
\title{From what country or countries did your ancestors come?  Spain}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethregion21
}
\details{
Question From what country or countries did your ancestors come?  Spain
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7660/vshow}{GSS Data Explorer page for \code{ethregion21}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 15 \tab 3167 \tab 135 \tab 681 \tab 1 \tab 33 \tab 4032 \cr
   2022 \tab 216 \tab 2980 \tab 137 \tab 163 \tab 48 \tab - \tab 3544 \cr
   Total \tab 231 \tab 6147 \tab 272 \tab 844 \tab 49 \tab 33 \tab 7576 \cr
}
}

\keyword{variable}
