% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_620.R
\name{ethregion68}
\alias{ethregion68}
\title{From what country or countries did your ancestors come?  Suriname}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethregion68
}
\details{
Question From what country or countries did your ancestors come?  Suriname
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7707/vshow}{GSS Data Explorer page for \code{ethregion68}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 15 \tab 3301 \tab 1 \tab 681 \tab 1 \tab 33 \tab 4032 \cr
   2022 \tab 216 \tab 3117 \tab - \tab 163 \tab 48 \tab - \tab 3544 \cr
   Total \tab 231 \tab 6418 \tab 1 \tab 844 \tab 49 \tab 33 \tab 7576 \cr
}
}

\keyword{variable}
