% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_621.R
\name{ethregion79}
\alias{ethregion79}
\title{From what country or countries did your ancestors come?  Panama}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethregion79
}
\details{
Question From what country or countries did your ancestors come?  Panama
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7718/vshow}{GSS Data Explorer page for \code{ethregion79}}.
}

\keyword{variable}
