% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_622.R
\name{ethregion95}
\alias{ethregion95}
\title{From what country or countries did your ancestors come?  Jamaica}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethregion95
}
\details{
Question From what country or countries did your ancestors come?  Jamaica
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7734/vshow}{GSS Data Explorer page for \code{ethregion95}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 15 \tab 3302 \tab 681 \tab 1 \tab 33 \tab 4032 \cr
   Total \tab 15 \tab 3302 \tab 681 \tab 1 \tab 33 \tab 4032 \cr
}
}

\keyword{variable}
