% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{ethregion96}
\alias{ethregion96}
\title{From what country or countries did your ancestors come?  Other oceania}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethregion96
}
\details{
Question From what country or countries did your ancestors come?  Other Oceania
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7735/vshow}{GSS Data Explorer page for \code{ethregion96}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 15 \tab 3297 \tab 5 \tab 681 \tab 1 \tab 33 \tab 4032 \cr
   2022 \tab 216 \tab 3111 \tab 6 \tab 163 \tab 48 \tab - \tab 3544 \cr
   Total \tab 231 \tab 6408 \tab 11 \tab 844 \tab 49 \tab 33 \tab 7576 \cr
}
}

\keyword{variable}
