% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{ethtrads}
\alias{ethtrads}
\title{Society recognize right of ethnic traditions}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethtrads
}
\details{
Question Here are some opinions some people have expressed in connection with ethnic issues in the United States. To what extent do you agree or disagree with each one?
C. If we want to help create a harmonious society, we must recognize that each ethnic group has the right to maintain its own unique traditions.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2728/vshow}{GSS Data Explorer page for \code{ethtrads}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 848 \tab 1430 \tab 266 \tab 147 \tab 39 \tab 28 \tab 7 \tab 2765 \cr
   Total \tab 0 \tab 848 \tab 1430 \tab 266 \tab 147 \tab 39 \tab 28 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethnofit}},
\code{\link{feelasns}},
\code{\link{feelblks}},
\code{\link{feelhsps}}
}
\concept{Prejudice}
\keyword{variable}
