% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_613.R
\name{ethworld7}
\alias{ethworld7}
\title{From what part of the world did your ancestors come?  North america}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethworld7
}
\details{
Question From what part of the world did your ancestors come?  North America
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7637/vshow}{GSS Data Explorer page for \code{ethworld7}}.
}

\keyword{variable}
