% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_613.R
\name{ethworld8}
\alias{ethworld8}
\title{From what part of the world did your ancestors come?  Caribbean}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethworld8
}
\details{
Question From what part of the world did your ancestors come?  Caribbean
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7638/vshow}{GSS Data Explorer page for \code{ethworld8}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab - \tab 3346 \tab 127 \tab 544 \tab 1 \tab 14 \tab 4032 \cr
   2022 \tab - \tab 3179 \tab 147 \tab 161 \tab 56 \tab 1 \tab 3544 \cr
   Total \tab 0 \tab 6525 \tab 274 \tab 705 \tab 57 \tab 15 \tab 7576 \cr
}
}

\keyword{variable}
