% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{evangcly}
\alias{evangcly}
\title{Is r evangelical christian}
\source{
General Social Survey https://gss.norc.org
}
\description{
evangcly
}
\details{
Question Would you describe yourself as an evangelical Christian, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7524/vshow}{GSS Data Explorer page for \code{evangcly}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 3240 \tab 289 \tab 477 \tab 6 \tab 1 \tab 17 \tab 2 \tab 4032 \cr
   Total \tab 3240 \tab 289 \tab 477 \tab 6 \tab 1 \tab 17 \tab 2 \tab 4032 \cr
}
}

\keyword{variable}
