% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{evbuyhme}
\alias{evbuyhme}
\title{Have you ever purchased a home}
\source{
General Social Survey https://gss.norc.org
}
\description{
evbuyhme
}
\details{
Question Have you ever purchased a home (including a condominium or mobile home)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1812/vshow}{GSS Data Explorer page for \code{evbuyhme}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   1996 \tab 2196 \tab 453 \tab 255 \tab 2904 \cr
   Total \tab 2196 \tab 453 \tab 255 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
