% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{evdrink}
\alias{evdrink}
\title{Ever drank any alcoholic beverage}
\source{
General Social Survey https://gss.norc.org
}
\description{
evdrink
}
\details{
Question Was there ever a time when you drank any alcoholic beverages?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3097/vshow}{GSS Data Explorer page for \code{evdrink}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 1996 \tab 669 \tab 139 \tab 8 \tab 2812 \cr
   Total \tab 1996 \tab 669 \tab 139 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkday}},
\code{\link{drinkmax}},
\code{\link{drinkyr}}
}
\concept{Women & Alcohol}
\keyword{variable}
