% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{evharjb}
\alias{evharjb}
\title{In life, r has ever been harassed on job bc r was believed to be gay/les/bi}
\source{
General Social Survey https://gss.norc.org
}
\description{
evharjb
}
\details{
Question Have you ever in your life been harassed verbally or in writing on the job by a co-worker, supervisor, or boss because you were believed to be gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3924/vshow}{GSS Data Explorer page for \code{evharjb}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1879 \tab 20 \tab 87 \tab 5 \tab 32 \tab 2023 \cr
   2012 \tab 1807 \tab 19 \tab 103 \tab - \tab 45 \tab 1974 \cr
   Total \tab 3686 \tab 39 \tab 190 \tab 5 \tab 77 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
