% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{evjob}
\alias{evjob}
\title{Have you ever had a paid job for 1 year or+}
\source{
General Social Survey https://gss.norc.org
}
\description{
evjob
}
\details{
Question Have you ever had a paid job for one year or more?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4593/vshow}{GSS Data Explorer page for \code{evjob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}

Other Unemployment: 
\code{\link{getjob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{yrjobend}}
}
\concept{Unemployment}
\concept{Work Orientation}
\keyword{variable}
