% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{evlosejb}
\alias{evlosejb}
\title{In life, r has ever been denied a job or fired bc employers believed r was gay/l}
\source{
General Social Survey https://gss.norc.org
}
\description{
evlosejb
}
\details{
Question Have you ever in your life been denied a job or been fired from a job because an employer believed you were gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3920/vshow}{GSS Data Explorer page for \code{evlosejb}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1879 \tab 11 \tab 96 \tab 5 \tab 32 \tab 2023 \cr
   2012 \tab 1807 \tab 7 \tab 114 \tab - \tab 46 \tab 1974 \cr
   Total \tab 3686 \tab 18 \tab 210 \tab 5 \tab 78 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
