% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{evolved}
\alias{evolved}
\title{Sci knowledge: human beings developed from animals}
\source{
General Social Survey https://gss.norc.org
}
\description{
evolved
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don't know or aren't sure, just tell me so,  and we will skip to the next question. Remember true, false, or don't know.
I. Human beings, as we know them today, developed from earlier species of animals. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3342/vshow}{GSS Data Explorer page for \code{evolved}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 779 \tab 791 \tab 290 \tab 4 \tab 4510 \cr
   2008 \tab 518 \tab 665 \tab 651 \tab 177 \tab 12 \tab 2023 \cr
   2010 \tab 1103 \tab 449 \tab 360 \tab 125 \tab 7 \tab 2044 \cr
   2012 \tab 1457 \tab 243 \tab 199 \tab 68 \tab 7 \tab 1974 \cr
   2014 \tab 1299 \tab 590 \tab 511 \tab 133 \tab 5 \tab 2538 \cr
   2016 \tab 2163 \tab 353 \tab 260 \tab 84 \tab 7 \tab 2867 \cr
   2018 \tab 1758 \tab 294 \tab 230 \tab 64 \tab 2 \tab 2348 \cr
   Total \tab 10944 \tab 3373 \tab 3002 \tab 941 \tab 44 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
