% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{evolved1}
\alias{evolved1}
\title{Humans developed from earlier species: true or false}
\source{
General Social Survey https://gss.norc.org
}
\description{
evolved1
}
\details{
Question According to the theory of evolution, human beings, as we know them today, developed from earlier species of animals. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3549/vshow}{GSS Data Explorer page for \code{evolved1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab can't choose \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 340 \tab 94 \tab 46 \tab 5 \tab 1974 \cr
   Total \tab 1489 \tab 340 \tab 94 \tab 46 \tab 5 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
