% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{evsmoke}
\alias{evsmoke}
\title{Has r ever smoked}
\source{
General Social Survey https://gss.norc.org
}
\description{
evsmoke
}
\details{
Question Do you smoke?
D. Have you ever smoked regularly?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/536/vshow}{GSS Data Explorer page for \code{evsmoke}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1978 \tab 604 \tab 300 \tab 573 \tab 55 \tab 1532 \cr
   1980 \tab 601 \tab 299 \tab 551 \tab 17 \tab 1468 \cr
   1983 \tab 591 \tab 329 \tab 667 \tab 12 \tab 1599 \cr
   1984 \tab 542 \tab 280 \tab 628 \tab 23 \tab 1473 \cr
   1986 \tab 477 \tab 301 \tab 648 \tab 44 \tab 1470 \cr
   1987 \tab 582 \tab 392 \tab 791 \tab 54 \tab 1819 \cr
   1988 \tab 823 \tab 200 \tab 408 \tab 50 \tab 1481 \cr
   1989 \tab 811 \tab 242 \tab 425 \tab 59 \tab 1537 \cr
   1990 \tab 758 \tab 220 \tab 368 \tab 26 \tab 1372 \cr
   1991 \tab 819 \tab 208 \tab 469 \tab 21 \tab 1517 \cr
   1993 \tab 837 \tab 260 \tab 496 \tab 13 \tab 1606 \cr
   1994 \tab 2623 \tab 124 \tab 239 \tab 6 \tab 2992 \cr
   Total \tab 10068 \tab 3155 \tab 6263 \tab 380 \tab 19866 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{cigweek}},
\code{\link{quitsmk}},
\code{\link{smokecig}}
}
\concept{Core}
\keyword{variable}
