% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{evunemp}
\alias{evunemp}
\title{Ever been unemployed two weeks +}
\source{
General Social Survey https://gss.norc.org
}
\description{
evunemp
}
\details{
Question Since your first full-time job have you ever been unemployed and looking for work for as long as two weeks?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[7]} never full-time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1272/vshow}{GSS Data Explorer page for \code{evunemp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab never full-time \tab no answer \tab Total \cr
   1991 \tab - \tab 686 \tab 755 \tab 69 \tab 7 \tab 1517 \cr
   Total \tab 0 \tab 686 \tab 755 \tab 69 \tab 7 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{numunemp}},
\code{\link{orgfin}},
\code{\link{totunemp}},
\code{\link{wrkyears}}
}
\concept{Work Organizations}
\keyword{variable}
