% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{exjobsat}
\alias{exjobsat}
\title{Satisfaction with past job}
\source{
General Social Survey https://gss.norc.org
}
\description{
exjobsat
}
\details{
Question (HAND CARD D20)
How satisfied were you in your last job?
}
\section{Values}{

\itemize{
\item \verb{[1]} completely satisfied
\item \verb{[2]} very satisfied
\item \verb{[3]} fairly satisfied
\item \verb{[4]} neither satisfied nor dissatisfied
\item \verb{[5]} fairly dissatisfied
\item \verb{[6]} very dissatisfied
\item \verb{[7]} completely dissatisfied
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6341/vshow}{GSS Data Explorer page for \code{exjobsat}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab completely satisfied \tab very satisfied \tab fairly satisfied \tab neither satisfied nor dissatisfied \tab fairly dissatisfied \tab very dissatisfied \tab completely dissatisfied \tab 99 \tab Total \cr
   2016 \tab 2382 \tab 134 \tab 147 \tab 109 \tab 21 \tab 15 \tab 23 \tab 13 \tab 23 \tab 2867 \cr
   Total \tab 2382 \tab 134 \tab 147 \tab 109 \tab 21 \tab 15 \tab 23 \tab 13 \tab 23 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
