% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_342.R
\name{expchng1}
\alias{expchng1}
\title{Cause of religious change1}
\source{
General Social Survey https://gss.norc.org
}
\description{
expchng1
}
\details{
Question Open-ended questions for how people who had undergone spiritual or religious change described their experiences. Occurrence or aspects of religious or spiritual change.
A. First mention
}
\section{Values}{

\itemize{
\item \verb{[96]} no mention
\item \verb{[97]} nothing special happened
\item \verb{[99]} no asnwer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3068/vshow}{GSS Data Explorer page for \code{expchng1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 4 \tab 40 \tab 41 \tab 44 \tab 48 \tab 49 \tab 5 \tab 52 \tab 54 \tab 55 \tab 56 \tab no mention \tab nothing special happened \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2163 \tab 5 \tab 3 \tab 6 \tab 1 \tab 2 \tab 2 \tab 6 \tab 12 \tab 4 \tab 1 \tab 2 \tab 3 \tab 25 \tab 3 \tab 8 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 5 \tab 2 \tab 2 \tab 2 \tab 6 \tab 8 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 469 \tab 5 \tab 3 \tab 36 \tab 2812 \cr
   Total \tab 2163 \tab 5 \tab 3 \tab 6 \tab 1 \tab 2 \tab 2 \tab 6 \tab 12 \tab 4 \tab 1 \tab 2 \tab 3 \tab 25 \tab 3 \tab 8 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 5 \tab 2 \tab 2 \tab 2 \tab 6 \tab 8 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 469 \tab 5 \tab 3 \tab 36 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{agerborn}},
\code{\link{chngrbrn}},
\code{\link{entity}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{numrborn}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}}
}
\concept{Spirituality and Religion}
\keyword{variable}
