% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_342.R
\name{expchng3}
\alias{expchng3}
\title{Cause of religious change3}
\source{
General Social Survey https://gss.norc.org
}
\description{
expchng3
}
\details{
Question Open-ended questions for how people who had undergone spiritual or religious change described their experiences. Occurrence or aspects of religious or spiritual change.
C. Third Mention
}
\section{Values}{

\itemize{
\item \verb{[96]} no mention
\item \verb{[97]} nothing special happened
\item \verb{[99]} no asnwer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3070/vshow}{GSS Data Explorer page for \code{expchng3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 2 \tab 54 \tab no mention \tab no answer \tab Total \cr
   2004 \tab 2163 \tab 1 \tab 1 \tab 611 \tab 36 \tab 2812 \cr
   Total \tab 2163 \tab 1 \tab 1 \tab 611 \tab 36 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{agerborn}},
\code{\link{chngrbrn}},
\code{\link{entity}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{numrborn}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}}
}
\concept{Spirituality and Religion}
\keyword{variable}
