% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{expdesgn}
\alias{expdesgn}
\title{Better way to test drug btw control and non-control}
\source{
General Social Survey https://gss.norc.org
}
\description{
expdesgn
}
\details{
Question Now, please think about this situation. Two scientists want to know if a certain drug is effective against high blood pressure. The first scientist wants to give the drug to one thousand people with high blood pressure and see how many of them experience lower blood pressure levels. The second scientist wants to give the drug to five hundred people with high blood pressure, and not give the drug to another five hundred people with high blood pressure, and see how many in both groups experience lower blood pressure levels. Which is the better way to test this drug?
}
\section{Values}{

\itemize{
\item \verb{[1]} all 1000 get the drug
\item \verb{[2]} 500 get the drug; 500 don't
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3330/vshow}{GSS Data Explorer page for \code{expdesgn}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab all 1000 get the drug \tab 500 get the drug; 500 don't \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 314 \tab 1476 \tab 70 \tab 4 \tab 4510 \cr
   2008 \tab 518 \tab 279 \tab 1115 \tab 93 \tab 18 \tab 2023 \cr
   2010 \tab 1328 \tab 99 \tab 571 \tab 36 \tab 10 \tab 2044 \cr
   2012 \tab 972 \tab 134 \tab 791 \tab 66 \tab 11 \tab 1974 \cr
   2014 \tab 1299 \tab 215 \tab 952 \tab 67 \tab 5 \tab 2538 \cr
   2016 \tab 1477 \tab 233 \tab 1098 \tab 49 \tab 10 \tab 2867 \cr
   2018 \tab 1173 \tab 222 \tab 903 \tab 39 \tab 11 \tab 2348 \cr
   Total \tab 9413 \tab 1496 \tab 6906 \tab 420 \tab 69 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{exptext}},
\code{\link{odds1}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
