% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{experts}
\alias{experts}
\title{Better to rely on opinions of experts than ordinary people}
\source{
General Social Survey https://gss.norc.org
}
\description{
experts
}
\details{
Question To what extent do you agree or disagree with the following statements?… It is better to rely on the opinions of experts than the experiences of ordinary people.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8048/vshow}{GSS Data Explorer page for \code{experts}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 97 \tab 338 \tab 502 \tab 450 \tab 134 \tab 101 \tab 1667 \tab 1 \tab 19 \tab 3309 \cr
   Total \tab 0 \tab 97 \tab 338 \tab 502 \tab 450 \tab 134 \tab 101 \tab 1667 \tab 1 \tab 19 \tab 3309 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{bizmoney}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{fulldem}},
\code{\link{immwlfare}},
\code{\link{incgap1}}
}
\concept{ISSP National Identity}
\keyword{variable}
