% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{exploit}
\alias{exploit}
\title{Other people take advantage of r if not careful}
\source{
General Social Survey https://gss.norc.org
}
\description{
exploit
}
\details{
Question To what extent do you agree or disagree with the following statements?
B. "If you are not careful, other people will take advantage of you." Do you...
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4118/vshow}{GSS Data Explorer page for \code{exploit}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 574 \tab 632 \tab 133 \tab 151 \tab 20 \tab 6 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 574 \tab 632 \tab 133 \tab 151 \tab 20 \tab 6 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{polsfair}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{wotrial}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
