% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{exptext}
\alias{exptext}
\title{Why is it better to test drug this way}
\source{
General Social Survey https://gss.norc.org
}
\description{
exptext
}
\details{
Question Why is it better to test the drug this way?
}
\section{Values}{

\itemize{
\item \code{1} correct - control group
\item \code{2} correct - vague reason
\item \code{3} correct - wrong reason
\item \code{4} correct - don't know
\item \code{5} wrong - reservations about control group
\item \code{6} wrong - reservations about sample
\item \code{7} wrong - other reason
\item \code{8} wrong - don't know
\item \code{9} no answer to closed ended question
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3331/vshow}{GSS Data Explorer page for \code{exptext}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{odds1}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
