% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{exptrain}
\alias{exptrain}
\title{Importance of different employer training for job}
\source{
General Social Survey https://gss.norc.org
}
\description{
exptrain
}
\details{
Question How important were each of the following as sources of the skills that you use in your job?  Would you say very important, important, somewhat important, or not at all important?
D. Experience or training at a different employer.
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} somewhat important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1371/vshow}{GSS Data Explorer page for \code{exptrain}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 196 \tab 256 \tab 166 \tab 275 \tab 8 \tab 11 \tab 1517 \cr
   Total \tab 605 \tab 196 \tab 256 \tab 166 \tab 275 \tab 8 \tab 11 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{baseofop}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{splocnum}},
\code{\link{training}},
\code{\link{wherewrk}}
}
\concept{Work Organizations}
\keyword{variable}
