% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{expunpop}
\alias{expunpop}
\title{Expression of unpopular ideas important for freedom}
\source{
General Social Survey https://gss.norc.org
}
\description{
expunpop
}
\details{
Question I’m going to read you several statements about freedom. Each one may say something true and important about freedom, but I’d like to know how important each statement is to you. Is it one of the most important things about freedom, extremely important, very important, somewhat important, or not too important?
F. Freedom is being able to express unpopular ideas without fearing for my safety.
}
\section{Values}{

\itemize{
\item \code{1} one of the most important
\item \code{2} extremely important
\item \code{3} very important
\item \code{4} moderately important
\item \code{5} somewhat important
\item \code{6} not too important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2673/vshow}{GSS Data Explorer page for \code{expunpop}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab one of the most important \tab extremely important \tab very important \tab moderately important \tab somewhat important \tab not too important \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 552 \tab 401 \tab 326 \tab 71 \tab 15 \tab 17 \tab 14 \tab 23 \tab 2817 \cr
   Total \tab 1398 \tab 552 \tab 401 \tab 326 \tab 71 \tab 15 \tab 17 \tab 14 \tab 23 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{cntrlife}},
\code{\link{freeprss}},
\code{\link{wlthpov}}
}
\concept{Freedom}
\keyword{variable}
