% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_379.R
\name{extinct}
\alias{extinct}
\title{Effect of global warming: polar bear's become extinct}
\source{
General Social Survey https://gss.norc.org
}
\description{
extinct
}
\details{
Question Please look at Card B24. Scientists predict that global warming may soon have big effects on the polar regions. I will describe some of these possible effects and, for each one, please say whether it would bother you a great deal, some, a little, or not at all if it actually happened.
A. By 2020, polar bears may become extinct.
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} some
\item \code{3} a little
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3438/vshow}{GSS Data Explorer page for \code{extinct}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{comorsci}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{nosun}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
