% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_321.R
\name{extr2001}
\alias{extr2001}
\title{Did r receive perf based payment in 2001}
\source{
General Social Survey https://gss.norc.org
}
\description{
extr2001
}
\details{
Question Did you receive any of these performance-based payments in the year 2001?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2859/vshow}{GSS Data Explorer page for \code{extr2001}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2152 \tab 395 \tab 200 \tab 1 \tab 17 \tab 2765 \cr
   Total \tab 2152 \tab 395 \tab 200 \tab 1 \tab 17 \tab 2765 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{deptperf}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{seecowrk}},
\code{\link{yearval}}
}
\concept{Employee Compensation}
\keyword{variable}
