% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_321.R
\name{extr2017}
\alias{extr2017}
\title{Did r receive any  performance-based incentives in 2017}
\source{
General Social Survey https://gss.norc.org
}
\description{
extr2017
}
\details{
Question Did you receive any of these performance-based payments in the year 2017?
IF ASKED: Performance-based pay means individual or group bonuses, commissions, or any type of profit-sharing.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7223/vshow}{GSS Data Explorer page for \code{extr2017}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2018 \tab 1848 \tab 360 \tab 140 \tab 2348 \cr
   Total \tab 1848 \tab 360 \tab 140 \tab 2348 \cr
}
}

\keyword{variable}
