% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{extr2021}
\alias{extr2021}
\title{Extra pay in 2021}
\source{
General Social Survey https://gss.norc.org
}
\description{
extr2021
}
\details{
Question Did you receive any of these performance-based payments in the year 2021? Performance-based pay means individual or group bonuses, commissions, or any type of profit-sharing.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7850/vshow}{GSS Data Explorer page for \code{extr2021}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2022 \tab 3222 \tab 207 \tab 103 \tab 10 \tab 2 \tab 3544 \cr
   Total \tab 3222 \tab 207 \tab 103 \tab 10 \tab 2 \tab 3544 \cr
}
}

\keyword{variable}
