% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_321.R
\name{extrapay}
\alias{extrapay}
\title{Eligible for performance based pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
extrapay
}
\details{
Question In your job are you eligible for any type of performance-based pay, such as individual or group bonuses or any type of profit-sharing?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2855/vshow}{GSS Data Explorer page for \code{extrapay}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{cowrkhrd}},
\code{\link{deptperf}},
\code{\link{donothng}},
\code{\link{extr2001}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{ownstock}},
\code{\link{seecowrk}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{workfor}},
\code{\link{workfor1}},
\code{\link{yearval}},
\code{\link{yearval1}}
}
\concept{Employee Compensation}
\keyword{variable}
