% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_321.R
\name{extraval}
\alias{extraval}
\title{Total dollar value of pay in previous year}
\source{
General Social Survey https://gss.norc.org
}
\description{
extraval
}
\details{
Question What was the approximate total dollar value of these various payments in 2001/5?
}
\section{Values}{

\itemize{
\item \code{9999995} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2860/vshow}{GSS Data Explorer page for \code{extraval}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{deptperf}},
\code{\link{extr2001}},
\code{\link{extrapay}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{seecowrk}},
\code{\link{yearval}}
}
\concept{Employee Compensation}
\keyword{variable}
