% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{exweathr}
\alias{exweathr}
\title{Neighborhood affected by extreme weather over the last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
exweathr
}
\details{
Question To what degree was your neighborhood affected by extreme weather over the last twelve months?
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} to a small extent
\item \code{3} to some extent
\item \code{4} to a great extent
\item \code{5} to a very great extent
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7503/vshow}{GSS Data Explorer page for \code{exweathr}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab not at all \tab to a small extent \tab to some extent \tab to a great extent \tab to a very great extent \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 301 \tab 554 \tab 591 \tab 212 \tab 98 \tab 64 \tab 201 \tab 8 \tab 4032 \cr
   Total \tab 2003 \tab 301 \tab 554 \tab 591 \tab 212 \tab 98 \tab 64 \tab 201 \tab 8 \tab 4032 \cr
}
}

\keyword{variable}
