% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_251.R
\name{fairblks}
\alias{fairblks}
\title{Blacks committed to fair and equal treatment of}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairblks
}
\details{
Question In the next statement a score of 1 means that you think almost all of the people in the group have a "commitment to the fair and equal treatment of all groups in society."  A score of 7 means that you think almost everyone in the group "lacks a commitment to the fair and equal treatment of all groups in society."  A score of 4 means that you think that the group is not towards one end or the other, and of course you may choose any number in between that comes closest to where you think people in the group stand.
B. Blacks?
}
\section{Values}{

\itemize{
\item \verb{[1]} commitment to fair and equal treatment
\item \verb{[7]} lacks commitment to fair and equal treatment
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2267/vshow}{GSS Data Explorer page for \code{fairblks}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab commitment to fair and equal treatment \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab lacks commitment to fair and equal treatment \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2000 \tab 1419 \tab 66 \tab 87 \tab 199 \tab 462 \tab 298 \tab 143 \tab 58 \tab 73 \tab 12 \tab - \tab 2817 \cr
   Total \tab 1419 \tab 66 \tab 87 \tab 199 \tab 462 \tab 298 \tab 143 \tab 58 \tab 73 \tab 12 \tab 0 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{betrlang}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blkwrk}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{ethchng}},
\code{\link{fairasns}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{knwwht}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{leastcom}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{uselang}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{uswht}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}

Other Blacks: 
\code{\link{acqblack}},
\code{\link{acqfmblk}},
\code{\link{acqnhblk}},
\code{\link{acqvablk}},
\code{\link{acqwkblk}},
\code{\link{admirblk}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkgovt}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blksimp}},
\code{\link{blkwrk}},
\code{\link{closeblk}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{comblk}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{famblks}},
\code{\link{fareblks}},
\code{\link{feelblks}},
\code{\link{influblk}},
\code{\link{intlblks}},
\code{\link{jobaff}},
\code{\link{knwblk}},
\code{\link{liveblks}},
\code{\link{marblk}},
\code{\link{patrblks}},
\code{\link{symptblk}},
\code{\link{trtblack}},
\code{\link{usblk}},
\code{\link{violblks}},
\code{\link{wlthblks}},
\code{\link{workblks}},
\code{\link{wrkwayup}}
}
\concept{Blacks}
\concept{Multi-Ethnic United States}
\keyword{variable}
