% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{fairdist}
\alias{fairdist}
\title{How fair is the income distribution in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairdist
}
\details{
Question How fair or unfair do you think the income distribution is in America?
}
\section{Values}{

\itemize{
\item \code{1} very fair
\item \code{2} fair
\item \code{3} unfair
\item \code{4} very unfair
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7515/vshow}{GSS Data Explorer page for \code{fairdist}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very fair \tab fair \tab unfair \tab very unfair \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 74 \tab 406 \tab 805 \tab 367 \tab 166 \tab 174 \tab 11 \tab 4032 \cr
   Total \tab 2029 \tab 74 \tab 406 \tab 805 \tab 367 \tab 166 \tab 174 \tab 11 \tab 4032 \cr
}
}

\keyword{variable}
