% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{fairearn}
\alias{fairearn}
\title{How fair is what r earn on the job}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairearn
}
\details{
Question How fair is what you earn on your job in comparison to others doing the same type of work you do?
}
\section{Values}{

\itemize{
\item \verb{[1]} much less than you deserve
\item \verb{[2]} somewhat less than you deserve
\item \verb{[3]} about as much as you deserve
\item \verb{[4]} somewhat more than you deserve
\item \verb{[5]} much more than you deserve
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2816/vshow}{GSS Data Explorer page for \code{fairearn}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab much less than you deserve \tab somewhat less than you deserve \tab about as much as you deserve \tab somewhat more than you deserve \tab much more than you deserve \tab don't know \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 969 \tab 252 \tab 475 \tab 874 \tab 116 \tab 25 \tab 37 \tab 17 \tab - \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 196 \tab 490 \tab 832 \tab 112 \tab 36 \tab 47 \tab 21 \tab - \tab - \tab 4510 \cr
   2010 \tab 857 \tab 165 \tab 350 \tab 534 \tab 62 \tab 20 \tab 25 \tab 31 \tab - \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 146 \tab 328 \tab 623 \tab 99 \tab 18 \tab 26 \tab 10 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 167 \tab 378 \tab 710 \tab 116 \tab 22 \tab 18 \tab 37 \tab - \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 234 \tab 610 \tab 909 \tab 127 \tab 28 \tab 36 \tab 20 \tab 25 \tab - \tab 3544 \cr
   Total \tab 8345 \tab 1160 \tab 2631 \tab 4482 \tab 632 \tab 149 \tab 189 \tab 136 \tab 25 \tab 0 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{chngtme}},
\code{\link{condemnd}},
\code{\link{cowrkhlp}},
\code{\link{cowrkint}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{fringeok}},
\code{\link{handmove}},
\code{\link{haveinfo}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{hlpequip}},
\code{\link{hlthdays}},
\code{\link{hrsrelax}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{hyperten}},
\code{\link{jobfind1}},
\code{\link{jobsecok}},
\code{\link{knowschd}},
\code{\link{knowschd1}},
\code{\link{knowwhat}},
\code{\link{laidoff}},
\code{\link{learnnew}},
\code{\link{lotofsay}},
\code{\link{manvsemp}},
\code{\link{misswork}},
\code{\link{mntlhlth}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{myskills}},
\code{\link{ntwkhard}},
\code{\link{opdevel}},
\code{\link{overwork}},
\code{\link{painarms}},
\code{\link{partteam}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{prodctiv}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{respect}},
\code{\link{rincblls}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{satjob1}},
\code{\link{secondwk}},
\code{\link{setthngs}},
\code{\link{slpprblm}},
\code{\link{spvtrfair}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{suphelp}},
\code{\link{teamsafe}},
\code{\link{toofewwk}},
\code{\link{trainops}},
\code{\link{trdunion}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{usedup}},
\code{\link{usetech}},
\code{\link{waypaid}},
\code{\link{weight}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbonus}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkpraise}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksmooth}},
\code{\link{wktopsat}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrktime}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Income: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{earndes}},
\code{\link{earnmore}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{famfinan}},
\code{\link{finalter}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{finrela}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{impfinan}},
\code{\link{incdef}},
\code{\link{inclstjb}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{justpay}},
\code{\link{minfood}},
\code{\link{minfour}},
\code{\link{mininc}},
\code{\link{minthree}},
\code{\link{othpay}},
\code{\link{owninc}},
\code{\link{pastpay}},
\code{\link{paychild}},
\code{\link{paychnge}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{payfam1}},
\code{\link{payhard}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{povline}},
\code{\link{realinc}},
\code{\link{realrinc}},
\code{\link{rincblls}},
\code{\link{rincom06}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{salsatfn}},
\code{\link{satfin}},
\code{\link{welfare1}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}}
}
\concept{Income}
\concept{Quality of Working Life}
\keyword{variable}
