% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{fairhwrk}
\alias{fairhwrk}
\title{Division of work unfair to you or spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairhwrk
}
\details{
Question We're interested in knowing how much of the work around your home is done by different people. We don't mean taking care of children - just regular work around the house, like cooking, grocery shopping and doing little repair jobs.
C. How fair do you feel the division of work around the house is in your household? Would you say that it is fair to both you and your spouse or partner, unfair to you, or unfair to your spouse or partner?
}
\section{Values}{

\itemize{
\item \code{1} fair to both
\item \code{2} unfair to me
\item \code{3} unfair to my spouse or partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1662/vshow}{GSS Data Explorer page for \code{fairhwrk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab fair to both \tab unfair to me \tab unfair to my spouse or partner \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2124 \tab 582 \tab 95 \tab 82 \tab 8 \tab 13 \tab 2904 \cr
   Total \tab 2124 \tab 582 \tab 95 \tab 82 \tab 8 \tab 13 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{earnmore}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}
}
\concept{Gender}
\keyword{variable}
