% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{fairnv}
\alias{fairnv}
\title{People fair or try to take advantage (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairnv
}
\details{
Question Do you think most people would try to take advantage of you if they got a chance, or would they try to be fair?
}
\section{Values}{

\itemize{
\item \verb{[1]} would take advantage of you
\item \verb{[2]} would try to be fair
\item \verb{[3]} depends (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7612/vshow}{GSS Data Explorer page for \code{fairnv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab would take advantage of you \tab would try to be fair \tab depends (phone mode only: volunteered) \tab skipped on web \tab Total \cr
   2021 \tab 3363 \tab 274 \tab 389 \tab 2 \tab 4 \tab 4032 \cr
   2022 \tab 2962 \tab 318 \tab 256 \tab 2 \tab 6 \tab 3544 \cr
   Total \tab 6325 \tab 592 \tab 645 \tab 4 \tab 10 \tab 7576 \cr
}
}

\keyword{variable}
