% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{fairv}
\alias{fairv}
\title{People fair or try to take advantage (with volunteered resposne on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairv
}
\details{
Question Do you think most people would try to take advantage of you if they got a chance, or would they try to be fair?
}
\section{Values}{

\itemize{
\item \verb{[1]} would take advantage of you
\item \verb{[2]} would try to be fair
\item \verb{[3]} depends (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7592/vshow}{GSS Data Explorer page for \code{fairv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab would take advantage of you \tab would try to be fair \tab depends (phone mode only: volunteered) \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 3344 \tab 140 \tab 280 \tab 264 \tab 1 \tab 3 \tab 4032 \cr
   2022 \tab 2974 \tab 156 \tab 148 \tab 262 \tab - \tab 4 \tab 3544 \cr
   Total \tab 6318 \tab 296 \tab 428 \tab 526 \tab 1 \tab 7 \tab 7576 \cr
}
}

\keyword{variable}
