% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{faith1}
\alias{faith1}
\title{Has death in family strengthened faith}
\source{
General Social Survey https://gss.norc.org
}
\description{
faith1
}
\details{
Question How often have these events strengthened your religious faith-often, sometimes, or never.
A. Death in the family
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1099/vshow}{GSS Data Explorer page for \code{faith1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 404 \tab 567 \tab 444 \tab 45 \tab 21 \tab 1481 \cr
   1998 \tab 1387 \tab 421 \tab 463 \tab 473 \tab 74 \tab 14 \tab 2832 \cr
   Total \tab 1387 \tab 825 \tab 1030 \tab 917 \tab 119 \tab 35 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
