% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_119.R
\name{faith3}
\alias{faith3}
\title{Has marriage strengthened faith}
\source{
General Social Survey https://gss.norc.org
}
\description{
faith3
}
\details{
Question How often have these events strengthened your religious faith-often, sometimes, or never.
C. Marriage
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1101/vshow}{GSS Data Explorer page for \code{faith3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 512 \tab 479 \tab 407 \tab 61 \tab 22 \tab 1481 \cr
   Total \tab 0 \tab 512 \tab 479 \tab 407 \tab 61 \tab 22 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{faith2}},
\code{\link{faith4}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}}
}
\concept{Religion}
\keyword{variable}
