% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{fambiz}
\alias{fambiz}
\title{What kind of business is this?}
\source{
General Social Survey https://gss.norc.org
}
\description{
fambiz
}
\details{
Question What kind of business is this?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} retail sales
\item \verb{[3]} wholesale sales
\item \verb{[4]} services
\item \verb{[5]} transportation
\item \verb{[6]} construction
\item \verb{[7]} finance, insrnce or real est
\item \verb{[8]} farming
\item \verb{[9]} other specify
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1878/vshow}{GSS Data Explorer page for \code{fambiz}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab manufacturing \tab retail sales \tab wholesale sales \tab services \tab transportation \tab construction \tab finance, insrnce or real est \tab farming \tab other specify \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 2841 \tab 4 \tab 6 \tab 1 \tab 11 \tab 5 \tab 9 \tab 1 \tab 6 \tab 8 \tab 1 \tab 11 \tab - \tab 2904 \cr
   Total \tab 2841 \tab 4 \tab 6 \tab 1 \tab 11 \tab 5 \tab 9 \tab 1 \tab 6 \tab 8 \tab 1 \tab 11 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{borrowed}},
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{econsys}},
\code{\link{evbuyhme}},
\code{\link{famfinan}},
\code{\link{fixer1}},
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{fixhome}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{haggle}},
\code{\link{homeyear}},
\code{\link{hrfambiz}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal3}},
\code{\link{inequal5}},
\code{\link{lawyer}},
\code{\link{lawyer1}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{layoffs}},
\code{\link{learnhme}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{natrecon}},
\code{\link{newowned}},
\code{\link{newused}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{realtor1}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{reldealr}},
\code{\link{relfixer}},
\code{\link{relhome}},
\code{\link{rellaw}},
\code{\link{relprivt}},
\code{\link{reqinfo}},
\code{\link{satcar}},
\code{\link{satfixes}},
\code{\link{sathome}},
\code{\link{satlawyr}},
\code{\link{sellauto}},
\code{\link{sellbaby}},
\code{\link{sellbed}},
\code{\link{sellhome}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{tagsales}},
\code{\link{transoth}},
\code{\link{transrel}},
\code{\link{typdealr}},
\code{\link{warrntyc}},
\code{\link{warrntyh}},
\code{\link{wholoand}},
\code{\link{whosold}},
\code{\link{wkfambiz}}

Other Agriculture: 
\code{\link{actfarm}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{givfarm}},
\code{\link{intfarm}},
\code{\link{memfarm}},
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{numfarm}},
\code{\link{payfarm}},
\code{\link{solfarm}},
\code{\link{typfarm}},
\code{\link{wkfambiz}},
\code{\link{yrfarm2}},
\code{\link{yrfarm3}}

Other Industry: 
\code{\link{indfirst}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{kdind80}},
\code{\link{maind80}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{sbind80}},
\code{\link{sphedind}},
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{wrkgovt}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{Agriculture}
\concept{Industry}
\concept{Markets}
\concept{Occupation}
\keyword{variable}
