% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{fambudgt}
\alias{fambudgt}
\title{How couples monitor budget}
\source{
General Social Survey https://gss.norc.org
}
\description{
fambudgt
}
\details{
Question How do you and your spouse/partner organize the income that one or both of you receive?  Please choose the option that comes closest.
}
\section{Values}{

\itemize{
\item \code{1} i manage, give prtnr share
\item \code{2} prntnr manages, gives me share
\item \code{3} pool all money, take out what we need
\item \code{4} pool some money, keep some separate
\item \code{5} keep money separate
\item \code{6} not married
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4417/vshow}{GSS Data Explorer page for \code{fambudgt}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}

Other Household: 
\code{\link{caresick}},
\code{\link{laundry}},
\code{\link{repairs}}
}
\concept{Household}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
