% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_612.R
\name{famdif16y}
\alias{famdif16y}
\title{Which of the following best describes your situation?}
\source{
General Social Survey https://gss.norc.org
}
\description{
famdif16y
}
\details{
Question Which of the following best describes your situation?
}
\section{Values}{

\itemize{
\item \code{1} one or both parents died
\item \code{2} parents divorced or separated
\item \code{3} parent absent in armed forces
\item \code{4} one or both parents in institution
\item \code{5} other (specify)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7629/vshow}{GSS Data Explorer page for \code{famdif16y}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab one or both parents died \tab parents divorced or separated \tab parent absent in armed forces \tab one or both parents in institution \tab other (specify) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 3768 \tab 85 \tab 147 \tab 1 \tab 3 \tab 28 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 2956 \tab 173 \tab 329 \tab 2 \tab 14 \tab 55 \tab 10 \tab 4 \tab 1 \tab 3544 \cr
   Total \tab 6724 \tab 258 \tab 476 \tab 3 \tab 17 \tab 83 \tab 10 \tab 4 \tab 1 \tab 7576 \cr
}
}

\keyword{variable}
