% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{famgen}
\alias{famgen}
\title{Number of family generations in household}
\source{
General Social Survey https://gss.norc.org
}
\description{
famgen
}
\details{
Question Number of family generations in household.
As of 2024, FAMGEN has been recoded to only account for 3+ generations. This has been retroactively applied to all previous years of FAMGEN. The original FAMGEN containing data 1975 to 2022 has been renamed FAMGEN_7522.
}
\section{Values}{

\itemize{
\item \code{1} one generation
\item \code{2} two generations, children
\item \code{3} two generations, parents
\item \code{4} two generations, grandchild(ren)
\item \code{5} three generations, child(ren), grandchildren
\item \code{6} three generations, child(ren), parent
\item \code{7} four generations
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5435/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab four generations \tab one generation \tab three generations, child(ren), grandchildren \tab three generations, child(ren), parent \tab two generations, children \tab two generations, grandchild(ren) \tab two generations, parents \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 1 \tab 615 \tab 31 \tab 21 \tab 797 \tab 8 \tab 17 \tab - \tab 1490 \cr
   1976 \tab 755 \tab 2 \tab 361 \tab 13 \tab 3 \tab 351 \tab 7 \tab 7 \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab 680 \tab 26 \tab 22 \tab 778 \tab 11 \tab 13 \tab - \tab 1530 \cr
   1978 \tab - \tab 1 \tab 703 \tab 24 \tab 13 \tab 768 \tab 12 \tab 11 \tab - \tab 1532 \cr
   1980 \tab - \tab 1 \tab 756 \tab 29 \tab 19 \tab 641 \tab 11 \tab 11 \tab - \tab 1468 \cr
   1982 \tab - \tab 2 \tab 928 \tab 60 \tab 15 \tab 810 \tab 24 \tab 21 \tab - \tab 1860 \cr
   1983 \tab - \tab 1 \tab 763 \tab 29 \tab 12 \tab 772 \tab 12 \tab 10 \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab 745 \tab 33 \tab 12 \tab 661 \tab 10 \tab 12 \tab - \tab 1473 \cr
   1985 \tab - \tab 1 \tab 799 \tab 17 \tab 19 \tab 674 \tab 7 \tab 17 \tab - \tab 1534 \cr
   1986 \tab - \tab 2 \tab 745 \tab 32 \tab 15 \tab 653 \tab 10 \tab 13 \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab 949 \tab 53 \tab 17 \tab 762 \tab 27 \tab 11 \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab 808 \tab 20 \tab 15 \tab 620 \tab 10 \tab 8 \tab - \tab 1481 \cr
   1989 \tab - \tab 1 \tab 843 \tab 37 \tab 10 \tab 632 \tab 9 \tab 5 \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab 751 \tab 21 \tab 8 \tab 573 \tab 9 \tab 10 \tab - \tab 1372 \cr
   1991 \tab - \tab 1 \tab 823 \tab 27 \tab 8 \tab 636 \tab 10 \tab 12 \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab 847 \tab 21 \tab 14 \tab 706 \tab 9 \tab 9 \tab - \tab 1606 \cr
   1994 \tab - \tab 1 \tab 1627 \tab 62 \tab 19 \tab 1240 \tab 19 \tab 24 \tab - \tab 2992 \cr
   1996 \tab - \tab 2 \tab 1662 \tab 54 \tab 29 \tab 1115 \tab 25 \tab 17 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab 1674 \tab 37 \tab 17 \tab 1070 \tab 12 \tab 22 \tab - \tab 2832 \cr
   2000 \tab - \tab 3 \tab 1589 \tab 56 \tab 27 \tab 1093 \tab 20 \tab 29 \tab - \tab 2817 \cr
   2002 \tab - \tab 1 \tab 1728 \tab 57 \tab 8 \tab 938 \tab 18 \tab 15 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab 1661 \tab 48 \tab 16 \tab 1029 \tab 23 \tab 35 \tab - \tab 2812 \cr
   2006 \tab - \tab 4 \tab 2617 \tab 66 \tab 34 \tab 1700 \tab 39 \tab 50 \tab - \tab 4510 \cr
   2008 \tab - \tab 2 \tab 1156 \tab 56 \tab 21 \tab 751 \tab 24 \tab 13 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab 1253 \tab 4 \tab 20 \tab 757 \tab 5 \tab 5 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 1187 \tab 2 \tab 23 \tab 741 \tab 1 \tab 20 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 1534 \tab 6 \tab 11 \tab 972 \tab 2 \tab 13 \tab - \tab 2538 \cr
   2016 \tab - \tab 1 \tab 1784 \tab 7 \tab 18 \tab 1032 \tab 9 \tab 16 \tab - \tab 2867 \cr
   2018 \tab - \tab 1 \tab 1478 \tab 53 \tab 30 \tab 738 \tab 18 \tab 30 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 2602 \tab 63 \tab 37 \tab 1216 \tab 44 \tab 70 \tab - \tab 4032 \cr
   2022 \tab 1764 \tab 4 \tab 1168 \tab 15 \tab 23 \tab 515 \tab 16 \tab 39 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 7120 \tab 32 \tab 36836 \tab 1059 \tab 556 \tab 25741 \tab 461 \tab 585 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{hhtype}},
\code{\link{hhtype1}},
\code{\link{relhh1}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}}

Other Household: 
\code{\link{hhtype}},
\code{\link{hhtype1}},
\code{\link{relhh1}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}}
}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
