% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_694.R
\name{famgen_exp_22}
\alias{famgen_exp_22}
\title{Number of family generations in household (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
famgen_exp_22
}
\details{
Question Recoded from experimental Household Composition variables ADULTS_EXP to RNRLTU18
As of 2024, FAMGEN_EXP has been recoded to only account for 3+ generations. This has been retroactively applied to all previous years of FAMGEN_EXP. The original FAMGEN containing data in 2022 has been renamed FAMGEN_EXP_22.
}
\section{Values}{

\itemize{
\item \verb{[1]} one generation
\item \verb{[2]} two generations, children
\item \verb{[3]} two generations, parents
\item \verb{[4]} two generations, grandchild(ren)
\item \verb{[5]} three generations, child(ren), grandchildren
\item \verb{[6]} three generations, child(ren), parent
\item \verb{[7]} four generations
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8233/vshow}{GSS Data Explorer page for \code{famgen_exp_22}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab one generation \tab two generations, children \tab two generations, parents \tab two generations, grandchild(ren) \tab three generations, child(ren), grandchildren \tab three generations, child(ren), parent \tab four generations \tab iap \tab Total \cr
   2022 \tab - \tab 1099 \tab 504 \tab 106 \tab 16 \tab 16 \tab 22 \tab 1 \tab 1780 \tab 3544 \cr
   Total \tab 0 \tab 1099 \tab 504 \tab 106 \tab 16 \tab 16 \tab 22 \tab 1 \tab 1780 \tab 3544 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{famgen_7522}},
\code{\link{rplace_7522}}

Other Household: 
\code{\link{famgen_7522}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp_22}},
\code{\link{rplace_7522}}
}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
