% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_690.R
\name{famgen_next}
\alias{famgen_next}
\title{Number of family generations in household - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
famgen_next
}
\details{
Question Given what you said, how many family generations live in this household?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} one generation
\item \verb{[2]} two generations, children
\item \verb{[3]} two generations, parents
\item \verb{[4]} two generations, grandchild(ren)
\item \verb{[5]} three or more generations
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8269/vshow}{GSS Data Explorer page for \code{famgen_next}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab one generation \tab two generations, children \tab two generations, parents \tab two generations, grandchild(ren) \tab three or more generations \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 670 \tab 235 \tab 68 \tab 11 \tab 31 \tab 2268 \tab 26 \tab 3309 \cr
   Total \tab 0 \tab 670 \tab 235 \tab 68 \tab 11 \tab 31 \tab 2268 \tab 26 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{hhtype1_next}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}}

Other GSS Next: 
\code{\link{hhtype1_next}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}}
}
\concept{GSS Next}
\keyword{variable}
