% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{family16sex}
\alias{family16sex}
\title{Sex of two parents living with when 16 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
family16sex
}
\details{
Question Were you living with your own two parents around the time you were 16? With whom were you living around that time?
}
\section{Values}{

\itemize{
\item \verb{[1]} mother and father
\item \verb{[2]} two same-sex parents
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8107/vshow}{GSS Data Explorer page for \code{family16sex}}.

Counts by year:\tabular{llllll}{
   year \tab not available in this year \tab mother and father \tab two same-sex parents \tab iap \tab Total \cr
   2024 \tab - \tab 2265 \tab 5 \tab 1039 \tab 3309 \cr
   Total \tab 0 \tab 2265 \tab 5 \tab 1039 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conorelgsp}},
\code{\link{femsee1}},
\code{\link{mascsee1}},
\code{\link{racdif5x}},
\code{\link{sexnow2}}

Other Sex: 
\code{\link{sexnow2}}
}
\concept{Core}
\concept{Sex}
\keyword{variable}
