% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_225.R
\name{fammhneg}
\alias{fammhneg}
\title{R's family's negative attitudes about mh problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
fammhneg
}
\details{
Question Thinking about your family, to what extent do they hold negative attitudes about people with mental health problems?
}
\section{Values}{

\itemize{
\item \verb{[1]} very much
\item \verb{[2]} somewhat
\item \verb{[3]} not very much
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7145/vshow}{GSS Data Explorer page for \code{fammhneg}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very much \tab somewhat \tab not very much \tab not at all \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2018 \tab 1175 \tab 31 \tab 192 \tab 390 \tab 477 \tab 31 \tab 52 \tab - \tab - \tab 2348 \cr
   2024 \tab 1693 \tab 116 \tab 370 \tab 572 \tab 496 \tab 32 \tab 9 \tab - \tab 21 \tab 3309 \cr
   Total \tab 2868 \tab 147 \tab 562 \tab 962 \tab 973 \tab 63 \tab 61 \tab 0 \tab 21 \tab 5657 \cr
}
}

\concept{Mental Health}
\keyword{variable}
