% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{famorjob}
\alias{famorjob}
\title{Would you sacrifice job opportunities for family}
\source{
General Social Survey https://gss.norc.org
}
\description{
famorjob
}
\details{
Question (HAND CARD D3)
Have you ever given up or would you give up good job opportunities for the benefit of your family life?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, i have done so and probably would do so again
\item \verb{[2]} yes, i have done so but probably would not do so again
\item \verb{[3]} no, i have not done so but probably would do so
\item \verb{[4]} no, i have not done so and probably would not do so
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6323/vshow}{GSS Data Explorer page for \code{famorjob}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, i have done so and probably would do so again \tab yes, i have done so but probably would not do so again \tab no, i have not done so but probably would do so \tab no, i have not done so and probably would not do so \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 620 \tab 117 \tab 568 \tab 154 \tab 9 \tab 9 \tab 2867 \cr
   Total \tab 1390 \tab 620 \tab 117 \tab 568 \tab 154 \tab 9 \tab 9 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
