% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_565.R
\name{famper2}
\alias{famper2}
\title{Separating from spouse or partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
famper2
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
E. What about family and personal relationships?  During the last year, did you
2. Separate from your husband/wife/partner.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5173/vshow}{GSS Data Explorer page for \code{famper2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab not available in this year \tab Total \cr
   1991 \tab 500 \tab 56 \tab 947 \tab 14 \tab - \tab - \tab 1517 \cr
   2004 \tab 1472 \tab 114 \tab 1213 \tab 12 \tab 1 \tab - \tab 2812 \cr
   Total \tab 1972 \tab 170 \tab 2160 \tab 26 \tab 1 \tab 0 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{extra1}},
\code{\link{extra2}},
\code{\link{famper1}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famper6}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth12}},
\code{\link{hlth13}},
\code{\link{hlth14}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}},
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}},
\code{\link{oth5}},
\code{\link{prob1}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{work1}},
\code{\link{work10}},
\code{\link{work2}},
\code{\link{work3}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}

Other Family: 
\code{\link{adoptkid}},
\code{\link{agewed}},
\code{\link{agewedcr}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{balwkfam}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{chldben}},
\code{\link{chlddth}},
\code{\link{chldhrt}},
\code{\link{clsclsex}},
\code{\link{cousins}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce}},
\code{\link{engda}},
\code{\link{engson}},
\code{\link{evdiv}},
\code{\link{evkid}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{famlife}},
\code{\link{famper1}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famvswk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{feovrwrk}},
\code{\link{flextime}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{inlawnum}},
\code{\link{inperson}},
\code{\link{jobvsfam}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidcall}},
\code{\link{kidlived}},
\code{\link{kidmost}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{kidvisit}},
\code{\link{letters}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{madeath}},
\code{\link{malive}},
\code{\link{malive1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{marital}},
\code{\link{marunion}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{meetings}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{mykids}},
\code{\link{niecenep}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{numcntct}},
\code{\link{numprobs}},
\code{\link{obtohelp}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{othkids}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{padeath}},
\code{\link{palive}},
\code{\link{palive1}},
\code{\link{parleave}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{pawrkgrw}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{posslq}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relcall}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relmarry}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{rfamlook}},
\code{\link{sibcall}},
\code{\link{sibdeath}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibvisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{spdeath}},
\code{\link{spfalook}},
\code{\link{spfam16}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{sppart}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{widowed}},
\code{\link{wkvsfam}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}}

Other Trauma: 
\code{\link{clsclsex}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{famper1}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}},
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}},
\code{\link{oth5}},
\code{\link{outofbiz}},
\code{\link{ownclerg}},
\code{\link{prob1}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unemp5}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{work1}},
\code{\link{work10}},
\code{\link{work2}},
\code{\link{work3}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}
}
\concept{Family}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
