% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_565.R
\name{famper3}
\alias{famper3}
\title{Breaking up with fianc?E)}
\source{
General Social Survey https://gss.norc.org
}
\description{
famper3
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
E. What about family and personal relationships?  During the last year, did you
3. Break up with a steady boyfriend/girlfriend or fiance(e).
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5174/vshow}{GSS Data Explorer page for \code{famper3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 500 \tab 52 \tab 958 \tab 7 \tab 1517 \cr
   2004 \tab 1472 \tab 124 \tab 1207 \tab 9 \tab 2812 \cr
   Total \tab 1972 \tab 176 \tab 2165 \tab 16 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{famper2}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famper6}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}}
}
\concept{Negative Life Events}
\keyword{variable}
