% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{fampress}
\alias{fampress}
\title{Do r family's members put pressure on r about the way r lives or organizes their}
\source{
General Social Survey https://gss.norc.org
}
\description{
fampress
}
\details{
Question In general, do your family members put pressure on you about the way you live or organize your personal life?
}
\section{Values}{

\itemize{
\item \code{1} no, never
\item \code{2} yes, but rarely
\item \code{3} yes, sometimes
\item \code{4} yes, often
\item \code{5} yes, very often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7259/vshow}{GSS Data Explorer page for \code{fampress}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab no, never \tab yes, but rarely \tab yes, sometimes \tab yes, often \tab yes, very often \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 626 \tab 239 \tab 211 \tab 54 \tab 38 \tab 4 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 626 \tab 239 \tab 211 \tab 54 \tab 38 \tab 4 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
