% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{famrnk}
\alias{famrnk}
\title{Rank in society of r}
\source{
General Social Survey https://gss.norc.org
}
\description{
famrnk
}
\details{
Question Now we would like to know where you and other people fit in. Please look at the diagram you think best describes America today, the one you chose first:
A. Where would you say you and your family actually are - at the top, somewhere in the middle, or toward the bottom?
}
\section{Values}{

\itemize{
\item \verb{[1]} top
\item \verb{[7]} bottom
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4338/vshow}{GSS Data Explorer page for \code{famrnk}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab top \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab bottom \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 26 \tab 33 \tab 197 \tab 487 \tab 248 \tab 93 \tab 51 \tab 54 \tab 83 \tab 2817 \cr
   Total \tab 1545 \tab 26 \tab 33 \tab 197 \tab 487 \tab 248 \tab 93 \tab 51 \tab 54 \tab 83 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
