% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{famsuffr}
\alias{famsuffr}
\title{Family life suffers if mom works f-t}
\source{
General Social Survey https://gss.norc.org
}
\description{
famsuffr
}
\details{
Question Do you agree or disagree ...
C. All in all, family life suffers when the woman has a full-time job.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1991, 1994, 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4345/vshow}{GSS Data Explorer page for \code{famsuffr}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 117 \tab 362 \tab 210 \tab 451 \tab 233 \tab 12 \tab 29 \tab - \tab 1481 \cr
   1991 \tab 158 \tab 123 \tab 329 \tab 227 \tab 391 \tab 232 \tab 17 \tab 40 \tab - \tab 1517 \cr
   1994 \tab 1545 \tab 133 \tab 367 \tab 196 \tab 504 \tab 226 \tab 7 \tab 14 \tab - \tab 2992 \cr
   1998 \tab 1548 \tab 136 \tab 244 \tab 240 \tab 338 \tab 273 \tab 26 \tab 27 \tab - \tab 2832 \cr
   2002 \tab 1594 \tab 158 \tab 299 \tab 263 \tab - \tab 434 \tab 15 \tab 2 \tab - \tab 2765 \cr
   2012 \tab 672 \tab 73 \tab 286 \tab 152 \tab 511 \tab 250 \tab 23 \tab 7 \tab - \tab 1974 \cr
   2022 \tab 2378 \tab 33 \tab 168 \tab 200 \tab 395 \tab 338 \tab 18 \tab 5 \tab 9 \tab 3544 \cr
   Total \tab 7962 \tab 773 \tab 2055 \tab 1488 \tab 2590 \tab 1986 \tab 118 \tab 124 \tab 9 \tab 17105 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fejobind}},
\code{\link{hapifwrk}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{kidsuffr}},
\code{\link{mawrkwrm}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
