% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_251.R
\name{famwhts}
\alias{famwhts}
\title{Whites committed to strong families}
\source{
General Social Survey https://gss.norc.org
}
\description{
famwhts
}
\details{
Question In the next statement a score of 1 means that you think almost all of the people in the group have a "commitment to strong families."  A score of 7 means that you think almost everyone in the group "lacks a commitment to strong families."  A score of 4 means that you think that the group is not towards one end or the other, and of course you may choose any number in between that comes closest to where you think people in the group stand.
A. Where would you rate Whites in general on this scale?
}
\section{Values}{

\itemize{
\item \code{1} commitment to strong families
\item \code{7} lacks commitment to strong families
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2261/vshow}{GSS Data Explorer page for \code{famwhts}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab commitment to strong families \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab lacks commitment to strong families \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 131 \tab 271 \tab 318 \tab 434 \tab 127 \tab 48 \tab 10 \tab 47 \tab 12 \tab 2817 \cr
   Total \tab 1419 \tab 131 \tab 271 \tab 318 \tab 434 \tab 127 \tab 48 \tab 10 \tab 47 \tab 12 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
