% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{famwklst}
\alias{famwklst}
\title{Wor'st way to org family work life with under school age child}
\source{
General Social Survey https://gss.norc.org
}
\description{
famwklst
}
\details{
Question And, in your opinion, which of these options would be the least desirable?
}
\section{Values}{

\itemize{
\item \code{1} the mother stays at home and the father works full-time
\item \code{2} the mother works part-time and the father works full-time
\item \code{3} both the mother and father work full-time
\item \code{4} both the mother and father work part-time
\item \code{5} the father works part-time and the mother works full-time
\item \code{6} the father stays at home and the mother works full-time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4468/vshow}{GSS Data Explorer page for \code{famwklst}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}

Other Family: 
\code{\link{famwkbst}}
}
\concept{Family}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
