% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_129.R
\name{fareasns}
\alias{fareasns}
\title{Self supporting - live off welfare}
\source{
General Social Survey https://gss.norc.org
}
\description{
fareasns
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
E. Do people in these groups tend to prefer to be self-supporting or do they tend to prefer to live off welfare?
4. Asian Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} self-supporting
\item \verb{[7]} live off welfare
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1205/vshow}{GSS Data Explorer page for \code{fareasns}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab self-supporting \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab live off welfare \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 122 \tab 209 \tab 217 \tab 426 \tab 150 \tab 58 \tab 23 \tab 137 \tab 30 \tab 1372 \cr
   Total \tab 0 \tab 122 \tab 209 \tab 217 \tab 426 \tab 150 \tab 58 \tab 23 \tab 137 \tab 30 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{intlhsps}},
\code{\link{intlso}},
\code{\link{patrjews}},
\code{\link{patrwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
