% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_129.R
\name{farewhts}
\alias{farewhts}
\title{Self supporting - live off welfare}
\source{
General Social Survey https://gss.norc.org
}
\description{
farewhts
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
E. Do people in these groups tend to prefer to be self-supporting or do they tend to prefer to live off welfare?
\enumerate{
\item Where would you rate whites in general on this scale?
}
}
\section{Values}{

\itemize{
\item \code{1} self-supporting
\item \code{7} live off welfare
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1202/vshow}{GSS Data Explorer page for \code{farewhts}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab self-supporting \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab live off welfare \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 210 \tab 394 \tab 345 \tab 294 \tab 36 \tab 17 \tab 6 \tab 40 \tab 30 \tab 1372 \cr
   Total \tab 0 \tab 210 \tab 394 \tab 345 \tab 294 \tab 36 \tab 17 \tab 6 \tab 40 \tab 30 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{intlhsps}},
\code{\link{intlso}},
\code{\link{patrjews}},
\code{\link{patrwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
