% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{fatalism}
\alias{fatalism}
\title{People can't change the course of their lives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fatalism
}
\details{
Question Do you agree or disagree with the following. . .
B. There is little that people can do to change the course of their lives.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4658/vshow}{GSS Data Explorer page for \code{fatalism}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 55 \tab 113 \tab 104 \tab 525 \tab 464 \tab 33 \tab 65 \tab 1517 \cr
   1998 \tab 1548 \tab 42 \tab 90 \tab 87 \tab 528 \tab 439 \tab 29 \tab 69 \tab 2832 \cr
   2008 \tab 658 \tab 47 \tab 107 \tab 78 \tab 618 \tab 506 \tab 7 \tab 2 \tab 2023 \cr
   2018 \tab 1173 \tab 46 \tab 103 \tab 84 \tab 472 \tab 460 \tab 3 \tab 7 \tab 2348 \cr
   Total \tab 3537 \tab 190 \tab 413 \tab 353 \tab 2143 \tab 1869 \tab 72 \tab 143 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{devil}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
