% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_38.R
\name{father}
\alias{father}
\title{Image of god as father}
\source{
General Social Survey https://gss.norc.org
}
\description{
father
}
\details{
Question When you think about God, how likely are each of these images to come to your mind?  Would you say extremely likely, somewhat likely, not too likely, or not likely at all? (HAND CARD AD. READ EACH ITEM.) (CIRCLE ONE NUMBER FOR EACH WORD)
G. Father
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not too likely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/328/vshow}{GSS Data Explorer page for \code{father}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab extremely likely \tab somewhat likely \tab not too likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab - \tab 950 \tab 379 \tab 108 \tab 113 \tab 30 \tab 19 \tab 1599 \cr
   1984 \tab 989 \tab 308 \tab 110 \tab 21 \tab 31 \tab 3 \tab 11 \tab 1473 \cr
   Total \tab 989 \tab 1258 \tab 489 \tab 129 \tab 144 \tab 33 \tab 30 \tab 3072 \cr
}
}

\seealso{
Other Core: 
\code{\link{creator}},
\code{\link{friend}},
\code{\link{judge}},
\code{\link{lover}},
\code{\link{master}},
\code{\link{mother}},
\code{\link{neargod}},
\code{\link{redeemer}},
\code{\link{spouse}}
}
\concept{Core}
\keyword{variable}
