% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_286.R
\name{fbaccess}
\alias{fbaccess}
\title{Permission to access r's facebook account}
\source{
General Social Survey https://gss.norc.org
}
\description{
fbaccess
}
\details{
Question Permission to access respondent's facebook account
}
\section{Values}{

\itemize{
\item \code{1} permission granted
\item \code{2} permission denied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7427/vshow}{GSS Data Explorer page for \code{fbaccess}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab permission granted \tab permission denied \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1903 \tab 317 \tab 623 \tab 2 \tab 22 \tab 2867 \cr
   Total \tab 1903 \tab 317 \tab 623 \tab 2 \tab 22 \tab 2867 \cr
}
}

\keyword{variable}
