% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{fearhome}
\alias{fearhome}
\title{Does r feel safe and secure at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
fearhome
}
\details{
Question How about at home at night--do you feel safe and secure, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/676/vshow}{GSS Data Explorer page for \code{fearhome}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 1554 \tab 277 \tab 11 \tab 18 \tab 1860 \cr
   Total \tab 0 \tab 1554 \tab 277 \tab 11 \tab 18 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{burglr}},
\code{\link{fear}},
\code{\link{owngun}},
\code{\link{pistol}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{rifle}},
\code{\link{robbry}},
\code{\link{shotgun}}
}
\concept{Core}
\keyword{variable}
