% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{febear}
\alias{febear}
\title{Can wife refuse to have children}
\source{
General Social Survey https://gss.norc.org
}
\description{
febear
}
\details{
Question If the husband in a family wants children, but the wife decides that she does not want any children, is it all right for the wife to refuse to have children?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/717/vshow}{GSS Data Explorer page for \code{febear}}.
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{feworkif}},
\code{\link{racsubgv}},
\code{\link{racsubs}}

Other Women: 
\code{\link{feworkif}}
}
\concept{Core}
\concept{Gender}
\concept{Parents}
\concept{Women}
\keyword{variable}
