% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{febrass}
\alias{febrass}
\title{Do women command over men}
\source{
General Social Survey https://gss.norc.org
}
\description{
febrass
}
\details{
Question C. As far as you know, are women in the armed forces now assigned to jobs where they have command over men, or are women not assigned to such jobs?  RECORD UNDER "C."
}
\section{Values}{

\itemize{
\item \code{1} yes, they are
\item \code{2} no they arent
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/800/vshow}{GSS Data Explorer page for \code{febrass}}.
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{fehlpmil}}
}
\concept{Core}
\keyword{variable}
