% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{fecare}
\alias{fecare}
\title{R's concern about women's rights}
\source{
General Social Survey https://gss.norc.org
}
\description{
fecare
}
\details{
Question A. How concerned are you personally about women''s rights?  Are you very concerned somewhat concerned, not very concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very concerned
\item \verb{[2]} somewhat concerned
\item \verb{[3]} not very concerned
\item \verb{[4]} not concerned at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/596/vshow}{GSS Data Explorer page for \code{fecare}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very concerned \tab somewhat concerned \tab not very concerned \tab not concerned at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab 804 \tab 171 \tab 377 \tab 162 \tab 69 \tab 11 \tab 5 \tab 1599 \cr
   Total \tab 804 \tab 171 \tab 377 \tab 162 \tab 69 \tab 11 \tab 5 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}
}
\concept{Core}
\keyword{variable}
