% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{feclergy}
\alias{feclergy}
\title{R favors women as ministers}
\source{
General Social Survey https://gss.norc.org
}
\description{
feclergy
}
\details{
Question B. ASK NON-CATHOLICS ONLY.
Do you favor or oppose women as pastors, ministers, priests, or rabbis in your own faith or denomination?
}
\section{Values}{

\itemize{
\item \verb{[1]} favor
\item \verb{[2]} oppose
\item \verb{[8]} no opinion
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/380/vshow}{GSS Data Explorer page for \code{feclergy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab favor \tab oppose \tab don't know \tab no answer \tab Total \cr
   1986 \tab 379 \tab 617 \tab 356 \tab 85 \tab 33 \tab 1470 \cr
   Total \tab 379 \tab 617 \tab 356 \tab 85 \tab 33 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}},
\code{\link{relgrade}}
}
\concept{Core}
\keyword{variable}
