% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{fecolop}
\alias{fecolop}
\title{College opportunities for women}
\source{
General Social Survey https://gss.norc.org
}
\description{
fecolop
}
\details{
Question Would you say that opportunities for college education are, in general, better or worse, for women than for men?
}
\section{Values}{

\itemize{
\item \code{1} much bettr for women
\item \code{2} better for women
\item \code{3} no difference
\item \code{4} worse for women
\item \code{5} much wrse for women
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4009/vshow}{GSS Data Explorer page for \code{fecolop}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much bettr for women \tab better for women \tab no difference \tab worse for women \tab much wrse for women \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 14 \tab 49 \tab 497 \tab 66 \tab 6 \tab 33 \tab 12 \tab 1534 \cr
   Total \tab 857 \tab 14 \tab 49 \tab 497 \tab 66 \tab 6 \tab 33 \tab 12 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}},
\code{\link{opfamily}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
