% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_88.R
\name{fedirty}
\alias{fedirty}
\title{Are women assigned to dirty jobs}
\source{
General Social Survey https://gss.norc.org
}
\description{
fedirty
}
\details{
Question B. As far as you know, are women in the armed forces now assigned to dirty jobs like repairing trucks or other heavy equipment, or are women not assigned to such jobs?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, they are
\item \verb{[2]} no they arent
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/799/vshow}{GSS Data Explorer page for \code{fedirty}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, they are \tab no they arent \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 863 \tab 429 \tab 546 \tab 22 \tab 1860 \cr
   Total \tab 0 \tab 863 \tab 429 \tab 546 \tab 22 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{brass}},
\code{\link{fefight}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{gunner}},
\code{\link{mechanic}},
\code{\link{nurse}},
\code{\link{transair}},
\code{\link{typist}}
}
\concept{Core}
\keyword{variable}
