% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{fedtrust}
\alias{fedtrust}
\title{How much does r trust federal government}
\source{
General Social Survey https://gss.norc.org
}
\description{
fedtrust
}
\details{
Question How much of the time do you think you can trust the government in Washington to do what is right -- just about always, most of the time only some of the time, or almost never?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} most of time
\item \verb{[3]} some of time
\item \verb{[4]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1038/vshow}{GSS Data Explorer page for \code{fedtrust}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost always \tab most of time \tab some of time \tab almost never \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1987 \tab - \tab 68 \tab 546 \tab 990 \tab 177 \tab 25 \tab 13 \tab - \tab 1819 \cr
   Total \tab 0 \tab 68 \tab 546 \tab 990 \tab 177 \tab 25 \tab 13 \tab 0 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{blkgains}},
\code{\link{blkinflu}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{gavepol}},
\code{\link{givegrp}},
\code{\link{governor}},
\code{\link{intpol}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{mempolit}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{mostact}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{swayvote}},
\code{\link{usrep}},
\code{\link{workpol}},
\code{\link{writegrp}}

Other Federal Government: 
\code{\link{advfront}},
\code{\link{affrmact}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{concong}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{congeth}},
\code{\link{congovt}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{demworks}},
\code{\link{equalize}},
\code{\link{govcheat}},
\code{\link{govtpow}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hlthcare}},
\code{\link{hspgovt}},
\code{\link{immcosts}},
\code{\link{immfare}},
\code{\link{jobaff}},
\code{\link{jobsall}},
\code{\link{knomentl}},
\code{\link{leadsci}},
\code{\link{mustret}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdiscm}},
\code{\link{natdisin}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natrecon}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{owneth}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{pricecon}},
\code{\link{reqinfo}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{scresrch}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secterr}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{spmentl}},
\code{\link{spnatdis}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{taxcheat}},
\code{\link{taxspend}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{whtgovt}}
}
\concept{Federal Government}
\concept{Socio-political participation}
\keyword{variable}
