% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{feelblks}
\alias{feelblks}
\title{How r feels towards african americans}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelblks
}
\details{
Question A. In general, how warm or cool do you feel towards African Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} very warm
\item \verb{[9]} very cool
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2730/vshow}{GSS Data Explorer page for \code{feelblks}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab very warm \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab very cool \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2002 \tab - \tab 740 \tab 318 \tab 364 \tab 264 \tab 745 \tab 105 \tab 99 \tab 35 \tab 59 \tab 24 \tab 12 \tab - \tab 2765 \cr
   Total \tab 0 \tab 740 \tab 318 \tab 364 \tab 264 \tab 745 \tab 105 \tab 99 \tab 35 \tab 59 \tab 24 \tab 12 \tab 0 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethdiff}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethnofit}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelhsps}},
\code{\link{feelwhts}},
\code{\link{whtsdiff}}

Other Blacks: 
\code{\link{acqblack}},
\code{\link{acqfmblk}},
\code{\link{acqnhblk}},
\code{\link{acqvablk}},
\code{\link{acqwkblk}},
\code{\link{admirblk}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkgovt}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blksimp}},
\code{\link{blkwrk}},
\code{\link{closeblk}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{comblk}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fairblks}},
\code{\link{famblks}},
\code{\link{fareblks}},
\code{\link{influblk}},
\code{\link{intlblks}},
\code{\link{jobaff}},
\code{\link{knwblk}},
\code{\link{liveblks}},
\code{\link{marblk}},
\code{\link{patrblks}},
\code{\link{symptblk}},
\code{\link{trtblack}},
\code{\link{usblk}},
\code{\link{violblks}},
\code{\link{wlthblks}},
\code{\link{workblks}},
\code{\link{wrkwayup}}
}
\concept{Blacks}
\concept{Prejudice}
\keyword{variable}
