% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{feeldown}
\alias{feeldown}
\title{How often felt down in past 2 weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
feeldown
}
\details{
Question Over the last 2 weeks, how often have you been bothered by the following problems: Feeling down, depressed, or hopeless.
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} several days
\item \code{3} more than half the days
\item \code{4} nearly every day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7861/vshow}{GSS Data Explorer page for \code{feeldown}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all \tab several days \tab more than half the days \tab nearly every day \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1557 \tab 1297 \tab 456 \tab 116 \tab 73 \tab 7 \tab 19 \tab 19 \tab 3544 \cr
   Total \tab 1557 \tab 1297 \tab 456 \tab 116 \tab 73 \tab 7 \tab 19 \tab 19 \tab 3544 \cr
}
}

\keyword{variable}
