% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{feelgod}
\alias{feelgod}
\title{Feel gods presence}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelgod
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
A. I feel God's presence.
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never/almst nevr
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2108/vshow}{GSS Data Explorer page for \code{feelgod}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never/almst nevr \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 245 \tab 361 \tab 195 \tab 193 \tab 237 \tab 167 \tab 18 \tab 29 \tab 2832 \cr
   2004 \tab 1472 \tab 296 \tab 347 \tab 186 \tab 148 \tab 166 \tab 169 \tab 9 \tab 19 \tab 2812 \cr
   Total \tab 2859 \tab 541 \tab 708 \tab 381 \tab 341 \tab 403 \tab 336 \tab 27 \tab 48 \tab 5644 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{relcmfrt}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
