% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{feelhsps}
\alias{feelhsps}
\title{How r feels towards hispanics}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelhsps
}
\details{
Question C. In general, how warm or cool do you feel towards Hispanics?
}
\section{Values}{

\itemize{
\item \code{1} very warm
\item \code{9} very cool
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2732/vshow}{GSS Data Explorer page for \code{feelhsps}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab very warm \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab very cool \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 633 \tab 349 \tab 372 \tab 290 \tab 697 \tab 121 \tab 133 \tab 43 \tab 72 \tab 45 \tab 10 \tab 2765 \cr
   Total \tab 0 \tab 633 \tab 349 \tab 372 \tab 290 \tab 697 \tab 121 \tab 133 \tab 43 \tab 72 \tab 45 \tab 10 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethnofit}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelblks}}
}
\concept{Prejudice}
\keyword{variable}
