% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{feelwhts}
\alias{feelwhts}
\title{How r feel towards caucasians}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelwhts
}
\details{
Question D. In general, how warm or cool do you feel towards white or Caucasian Americans?
}
\section{Values}{

\itemize{
\item \code{1} very warm
\item \code{9} very cool
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2733/vshow}{GSS Data Explorer page for \code{feelwhts}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab very warm \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab very cool \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 953 \tab 477 \tab 361 \tab 245 \tab 522 \tab 55 \tab 53 \tab 26 \tab 42 \tab 21 \tab 10 \tab 2765 \cr
   Total \tab 0 \tab 953 \tab 477 \tab 361 \tab 245 \tab 522 \tab 55 \tab 53 \tab 26 \tab 42 \tab 21 \tab 10 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethdiff}},
\code{\link{ethsame}},
\code{\link{whtsdiff}}
}
\concept{Prejudice}
\keyword{variable}
