% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_88.R
\name{fefight}
\alias{fefight}
\title{Are women assigned to combat}
\source{
General Social Survey https://gss.norc.org
}
\description{
fefight
}
\details{
Question A. As far as you know, are women now assigned to jobs in the armed forces that would expose them to combat, or are women not assigned to such jobs?  RECORD UNDER "A" BELOW.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, they are
\item \verb{[2]} no they arent
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/798/vshow}{GSS Data Explorer page for \code{fefight}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, they are \tab no they arent \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 418 \tab 960 \tab 462 \tab 20 \tab 1860 \cr
   Total \tab 0 \tab 418 \tab 960 \tab 462 \tab 20 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{brass}},
\code{\link{fedirty}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{gunner}},
\code{\link{mechanic}},
\code{\link{nurse}},
\code{\link{transair}},
\code{\link{typist}}
}
\concept{Core}
\keyword{variable}
