% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{fefirm}
\alias{fefirm}
\title{How firm is r's opinion on women's rghts}
\source{
General Social Survey https://gss.norc.org
}
\description{
fefirm
}
\details{
Question C. How firm are you about your opinion on women''s rights--would you say you are very likely to change your opinion, somewhat likely to change, somewhat unlikely to change, or very unlikely to change?
}
\section{Values}{

\itemize{
\item \code{1} change very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} change very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/595/vshow}{GSS Data Explorer page for \code{fefirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab change very likely \tab somewhat likely \tab somewhat unlikely \tab change very unlikely \tab don't know \tab no answer \tab Total \cr
   1983 \tab 795 \tab 36 \tab 243 \tab 223 \tab 273 \tab 25 \tab 4 \tab 1599 \cr
   Total \tab 795 \tab 36 \tab 243 \tab 223 \tab 273 \tab 25 \tab 4 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{fecare}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}
}
\concept{Core}
\keyword{variable}
