% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{fegive}
\alias{fegive}
\title{Has r given money to women's rights orgs}
\source{
General Social Survey https://gss.norc.org
}
\description{
fegive
}
\details{
Question Have you ever . . .
B. Given money to an organization concerned with this issue?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/599/vshow}{GSS Data Explorer page for \code{fegive}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1983 \tab 804 \tab 88 \tab 701 \tab 1 \tab 5 \tab 1599 \cr
   Total \tab 804 \tab 88 \tab 701 \tab 1 \tab 5 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}
}
\concept{Core}
\keyword{variable}
