% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{fehelp}
\alias{fehelp}
\title{Wife should help husbands career first}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehelp
}
\details{
Question Now I'm going to read several more statements. As I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, here is the statement:
B. It is more important for a wife to help her husband's career than to have one herself.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1998 \cr
   A/B/C \tab 1996 \cr
   Full \tab 1977, 1985, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/704/vshow}{GSS Data Explorer page for \code{fehelp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 205 \tab 636 \tab 534 \tab 97 \tab 52 \tab 6 \tab 1530 \cr
   1985 \tab - \tab 97 \tab 455 \tab 679 \tab 241 \tab 55 \tab 7 \tab 1534 \cr
   1986 \tab - \tab 88 \tab 428 \tab 681 \tab 239 \tab 33 \tab 1 \tab 1470 \cr
   1988 \tab 493 \tab 47 \tab 253 \tab 466 \tab 200 \tab 21 \tab 1 \tab 1481 \cr
   1989 \tab 531 \tab 53 \tab 217 \tab 501 \tab 194 \tab 37 \tab 4 \tab 1537 \cr
   1990 \tab 444 \tab 35 \tab 230 \tab 457 \tab 182 \tab 22 \tab 2 \tab 1372 \cr
   1991 \tab 493 \tab 34 \tab 262 \tab 504 \tab 200 \tab 23 \tab 1 \tab 1517 \cr
   1993 \tab 526 \tab 33 \tab 214 \tab 576 \tab 222 \tab 33 \tab 2 \tab 1606 \cr
   1994 \tab 1015 \tab 59 \tab 351 \tab 1057 \tab 433 \tab 64 \tab 13 \tab 2992 \cr
   1996 \tab 475 \tab 74 \tab 420 \tab 1278 \tab 574 \tab 73 \tab 10 \tab 2904 \cr
   1998 \tab 961 \tab 44 \tab 303 \tab 999 \tab 459 \tab 54 \tab 12 \tab 2832 \cr
   Total \tab 4938 \tab 769 \tab 3769 \tab 7732 \tab 3041 \tab 467 \tab 59 \tab 20775 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{ablegal}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
