% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{fehlpmil}
\alias{fehlpmil}
\title{Do women raise effectiveness of military}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehlpmil
}
\details{
Question In the past 10 years the number of women in the armed forces has increased from about 2 percent to about 8 percent
today. In general, would you say the increased number of women has raised the effectiveness of our armed forces, has it made no difference, or has it made our armed forces less effective?
(Just your own opinion.)
}
\section{Values}{

\itemize{
\item \verb{[1]} raised effectiveness
\item \verb{[2]} no difference
\item \verb{[3]} lowered effctiveness
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/801/vshow}{GSS Data Explorer page for \code{fehlpmil}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab raised effectiveness \tab no difference \tab lowered effctiveness \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 364 \tab 1119 \tab 145 \tab 221 \tab 11 \tab 1860 \cr
   Total \tab 0 \tab 364 \tab 1119 \tab 145 \tab 221 \tab 11 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}}
}
\concept{Core}
\keyword{variable}
