% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{feimp}
\alias{feimp}
\title{Importance of women's rights issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
feimp
}
\details{
Question A. How important is the women''s rights issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most imp
\item \verb{[2]} important
\item \verb{[3]} not very imp
\item \verb{[4]} not imp at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   Full \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/593/vshow}{GSS Data Explorer page for \code{feimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not imp at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab 795 \tab 77 \tab 389 \tab 231 \tab 93 \tab 9 \tab 5 \tab 1599 \cr
   1996 \tab 1444 \tab 164 \tab 760 \tab 366 \tab 144 \tab 24 \tab 2 \tab 2904 \cr
   Total \tab 2239 \tab 241 \tab 1149 \tab 597 \tab 237 \tab 33 \tab 7 \tab 4503 \cr
}
}

\seealso{
Other Core: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}
}
\concept{Core}
\concept{Gender}
\keyword{variable}
