% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{feinflu}
\alias{feinflu}
\title{Opinion of political influence of women}
\source{
General Social Survey https://gss.norc.org
}
\description{
feinflu
}
\details{
Question People have different opinions about the amount of influence that various groups have in American life and politics. Do you think that women have far too much influence, too much influence, about the right amount of influence, too little influence, or do they have far too little influence?
}
\section{Values}{

\itemize{
\item \verb{[1]} far too much
\item \verb{[2]} too much
\item \verb{[3]} right amount
\item \verb{[4]} too little
\item \verb{[5]} far too little
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1043/vshow}{GSS Data Explorer page for \code{feinflu}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab far too much \tab too much \tab right amount \tab too little \tab far too little \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1987 \tab 778 \tab 8 \tab 43 \tab 375 \tab 461 \tab 118 \tab 25 \tab 11 \tab - \tab 1819 \cr
   Total \tab 778 \tab 8 \tab 43 \tab 375 \tab 461 \tab 118 \tab 25 \tab 11 \tab 0 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{blkgains}},
\code{\link{blkinflu}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{fedtrust}},
\code{\link{fegains}},
\code{\link{gavepol}},
\code{\link{givegrp}},
\code{\link{governor}},
\code{\link{intpol}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{mempolit}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{mostact}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{swayvote}},
\code{\link{usrep}},
\code{\link{workpol}},
\code{\link{writegrp}}

Other Women: 
\code{\link{children}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{earnmore}},
\code{\link{famsuffr}},
\code{\link{febear}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fecolop}},
\code{\link{fefam}},
\code{\link{fefirm}},
\code{\link{fegains}},
\code{\link{fehelp}},
\code{\link{fehire}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejobaff}},
\code{\link{fejobind}},
\code{\link{fejobop}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{fewknokd}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrksup}},
\code{\link{hapifwrk}},
\code{\link{hmemaker}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{kidsuffr}},
\code{\link{manprof}},
\code{\link{mawrkwrm}},
\code{\link{mebear}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{men}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{opsex}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{salfergt}},
\code{\link{sexpromo}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{wrkbaby}},
\code{\link{wrkclass}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}},
\code{\link{yourself}}
}
\concept{Socio-political participation}
\concept{Women}
\keyword{variable}
