% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{feinfo}
\alias{feinfo}
\title{How much info r has on women's rgts}
\source{
General Social Survey https://gss.norc.org
}
\description{
feinfo
}
\details{
Question B. How much information do you have about the women's rights issue? Do you have all of the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \code{1} all info r needs
\item \code{2} most info
\item \code{3} some info
\item \code{4} very little info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/594/vshow}{GSS Data Explorer page for \code{feinfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all info r needs \tab most info \tab some info \tab very little info \tab don't know \tab no answer \tab Total \cr
   1983 \tab 795 \tab 70 \tab 172 \tab 323 \tab 227 \tab 8 \tab 4 \tab 1599 \cr
   Total \tab 795 \tab 70 \tab 172 \tab 323 \tab 227 \tab 8 \tab 4 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}
}
\concept{Core}
\keyword{variable}
