% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{fejobind}
\alias{fejobind}
\title{Job is best way for woman to be indep}
\source{
General Social Survey https://gss.norc.org
}
\description{
fejobind
}
\details{
Question Do you agree or disagree ...
G. Having a job is the best way for a woman to be an independent person.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4349/vshow}{GSS Data Explorer page for \code{fejobind}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 157 \tab 444 \tab 316 \tab 355 \tab 82 \tab 37 \tab 23 \tab 1481 \cr
   1994 \tab 1545 \tab 175 \tab 464 \tab 309 \tab 374 \tab 86 \tab 21 \tab 18 \tab 2992 \cr
   2002 \tab 1594 \tab 257 \tab 351 \tab 291 \tab - \tab 254 \tab 17 \tab 1 \tab 2765 \cr
   Total \tab 3206 \tab 589 \tab 1259 \tab 916 \tab 729 \tab 422 \tab 75 \tab 42 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famsuffr}},
\code{\link{hapifwrk}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{kidsuffr}},
\code{\link{mawrkwrm}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
