% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{fejobop}
\alias{fejobop}
\title{Job opportunities for women}
\source{
General Social Survey https://gss.norc.org
}
\description{
fejobop
}
\details{
Question How about job opportunities for women -- do you think they are, in general, better or worse than job opportunities for men with similar education and experience?
}
\section{Values}{

\itemize{
\item \verb{[1]} much bettr for women
\item \verb{[2]} better for women
\item \verb{[3]} no difference
\item \verb{[4]} worse for women
\item \verb{[5]} much wrse for women
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4010/vshow}{GSS Data Explorer page for \code{fejobop}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much bettr for women \tab better for women \tab no difference \tab worse for women \tab much wrse for women \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 18 \tab 55 \tab 212 \tab 321 \tab 31 \tab 30 \tab 10 \tab 1534 \cr
   Total \tab 857 \tab 18 \tab 55 \tab 212 \tab 321 \tab 31 \tab 30 \tab 10 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}},
\code{\link{opfamily}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
