% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_180.R
\name{fekids1}
\alias{fekids1}
\title{How important women biologically better-suited}
\source{
General Social Survey https://gss.norc.org
}
\description{
fekids1
}
\details{
Question Women are more likely than men to take care of children. I'm going to read several reasons why this might be so. please tell me how important you think each reason is: very important, important, somewhat important, or not at all important. For example, ...
A. Women are biologically better-suited to care for children. How important do you think this reason is for explaining why women are more likely to take care of children?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} somewhat important
\item \code{4} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1691/vshow}{GSS Data Explorer page for \code{fekids1}}.
}

\seealso{
Other Gender: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{womenhrt}}
}
\concept{Gender}
\keyword{variable}
