% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_180.R
\name{feless1}
\alias{feless1}
\title{How important men work harder explain}
\source{
General Social Survey https://gss.norc.org
}
\description{
feless1
}
\details{
Question On average, women who are employed full time earn less than men earn. I'm going to read several reasons why this might be so. Please tell me how important you think each reason is: very important, important, somewhat important, or not at all important. For example,...
A. Men work harder on the job than women Do. How important do you think this reason is for explaining why women earn less?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} somewhat important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1688/vshow}{GSS Data Explorer page for \code{feless1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 125 \tab 170 \tab 200 \tab 902 \tab 52 \tab 11 \tab 2904 \cr
   Total \tab 1444 \tab 125 \tab 170 \tab 200 \tab 902 \tab 52 \tab 11 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{fekids1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{womenhrt}}
}
\concept{Gender}
\keyword{variable}
