% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{femarry}
\alias{femarry}
\title{Way of life recommendation for woman}
\source{
General Social Survey https://gss.norc.org
}
\description{
femarry
}
\details{
Question If you were advising a young woman, which of the following ways of life would you recommend?
}
\section{Values}{

\itemize{
\item \verb{[1]} live alone
\item \verb{[2]} live with posslq
\item \verb{[3]} cohabitate, then marry
\item \verb{[4]} marry
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4362/vshow}{GSS Data Explorer page for \code{femarry}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab live alone \tab live with posslq \tab cohabitate, then marry \tab marry \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 131 \tab 37 \tab 372 \tab 652 \tab 197 \tab 25 \tab 1481 \cr
   Total \tab 67 \tab 131 \tab 37 \tab 372 \tab 652 \tab 197 \tab 25 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
