% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{feminist}
\alias{feminist}
\title{Think yourself as a feminist?}
\source{
General Social Survey https://gss.norc.org
}
\description{
feminist
}
\details{
Question Do you think of yourself as a feminist or not?
}
\section{Values}{

\itemize{
\item \code{1} yes, a feminist
\item \code{2} no, not a feminist
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1698/vshow}{GSS Data Explorer page for \code{feminist}}.
}

\seealso{
Other Gender: 
\code{\link{fehire}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{fenews}},
\code{\link{hmemaker}},
\code{\link{mebear}},
\code{\link{wrkclass}}
}
\concept{Gender}
\keyword{variable}
