% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{femself}
\alias{femself}
\title{How feminine r sees self}
\source{
General Social Survey https://gss.norc.org
}
\description{
femself
}
\details{
Question How femininte do you see yourself?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{7} very
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7960/vshow}{GSS Data Explorer page for \code{femself}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very \tab Total \cr
   2022 \tab 2504 \tab 205 \tab 113 \tab 45 \tab 59 \tab 138 \tab 132 \tab 348 \tab 3544 \cr
   Total \tab 2504 \tab 205 \tab 113 \tab 45 \tab 59 \tab 138 \tab 132 \tab 348 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{firstkidsex}},
\code{\link{mascself}},
\code{\link{nonbinkids}}
}
\concept{GSS Next}
\keyword{variable}
