% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{fenumok}
\alias{fenumok}
\title{Number of women in military}
\source{
General Social Survey https://gss.norc.org
}
\description{
fenumok
}
\details{
Question As you know, this country stopped the military draft in 1972. Since that time we have relied on volunteers. Now I'd like to ask you a few questions about our armed forces.
C. At the present time, about 9 percent of the armed forces are women. All things considered, do you think there are too many women in the armed forces, about the right number, or should there be more women in the armed forces?
RECORD UNDER "C" BELOW.
}
\section{Values}{

\itemize{
\item \code{1} too many
\item \code{2} about right number
\item \code{3} should be more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/783/vshow}{GSS Data Explorer page for \code{fenumok}}.
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
