% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{feovrwrk}
\alias{feovrwrk}
\title{Family life suffers when women work}
\source{
General Social Survey https://gss.norc.org
}
\description{
feovrwrk
}
\details{
Question Do you agree or disagree? Family life often suffers because women concentrate too much on their work.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8218/vshow}{GSS Data Explorer page for \code{feovrwrk}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 125 \tab 395 \tab 722 \tab 620 \tab 229 \tab 29 \tab 1162 \tab 4 \tab 23 \tab 3309 \cr
   Total \tab 0 \tab 125 \tab 395 \tab 722 \tab 620 \tab 229 \tab 29 \tab 1162 \tab 4 \tab 23 \tab 3309 \cr
}
}

\concept{Family}
\keyword{variable}
