% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{fepolnv}
\alias{fepolnv}
\title{Women not suited for politics (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fepolnv
}
\details{
Question Please indicate if you agree or disagree with this statement: Most men are better suited emotionally for politics than are most women.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[3]} not sure (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7622/vshow}{GSS Data Explorer page for \code{fepolnv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2689 \tab 151 \tab 1179 \tab 5 \tab 1 \tab 7 \tab 4032 \cr
   2022 \tab 2944 \tab 72 \tab 524 \tab - \tab - \tab 4 \tab 3544 \cr
   Total \tab 5633 \tab 223 \tab 1703 \tab 5 \tab 1 \tab 11 \tab 7576 \cr
}
}

\keyword{variable}
