% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{fepolv}
\alias{fepolv}
\title{Women not suited for politics (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fepolv
}
\details{
Question Please indicate if you agree or disagree with this statement: Most men are better suited emotionally for politics than are most women.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[3]} not sure (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7602/vshow}{GSS Data Explorer page for \code{fepolv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2658 \tab 126 \tab 1104 \tab 140 \tab 4 \tab 4032 \cr
   2022 \tab 2990 \tab 51 \tab 435 \tab 68 \tab - \tab 3544 \cr
   2024 \tab 2737 \tab 81 \tab 419 \tab 72 \tab - \tab 3309 \cr
   Total \tab 8385 \tab 258 \tab 1958 \tab 280 \tab 4 \tab 10885 \cr
}
}

\keyword{variable}
