% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{fepriest}
\alias{fepriest}
\title{R favors women as priests}
\source{
General Social Survey https://gss.norc.org
}
\description{
fepriest
}
\details{
Question A. IF CATHOLIC RESPONDENT, ASK FEPRIEST. OTHERS GO TO FECLERGY.
As I read this statement, will you tell me whether you agree strongly, agree somewhat, disagree somewhat, or disagree strongly?
It would be a good thing if women were allowed to be ordained as priests.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} disagree strongly
\item \code{8} no opinion
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/379/vshow}{GSS Data Explorer page for \code{fepriest}}.
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}},
\code{\link{relgrade}}

Other Religion: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{relgrade}}

Other Women: 
\code{\link{feclergy}}
}
\concept{Core}
\concept{Religion}
\concept{Women}
\keyword{variable}
