% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{ferespct}
\alias{ferespct}
\title{How well local church respects women's rights}
\source{
General Social Survey https://gss.norc.org
}
\description{
ferespct
}
\details{
Question How do you rate how your local church does each of the following?  Would you say your local church does an excellent, very good, fair, or poor job of:
B. Respecting the Rights of Women
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1162/vshow}{GSS Data Explorer page for \code{ferespct}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1831 \tab 289 \tab 278 \tab 208 \tab 57 \tab 13 \tab 67 \tab 74 \tab 2817 \cr
   Total \tab 1831 \tab 289 \tab 278 \tab 208 \tab 57 \tab 13 \tab 67 \tab 74 \tab 2817 \cr
}
}

\seealso{
Other Religion: 
\code{\link{obeytch}},
\code{\link{preach}},
\code{\link{sympcoun}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
