% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{feserve}
\alias{feserve}
\title{Mandatory conscription for females?}
\source{
General Social Survey https://gss.norc.org
}
\description{
feserve
}
\details{
Question A. How would you feel about a program that required all young women to give one year of service to the nation--either in the military forces or in non-military work such as in hospitals or with elderly people--Would you strongly favor it, probably favor it, probably oppose it, or strongly oppose it?
RECORD UNDER "A"
BELOW.
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} probably favor
\item \code{3} probably oppose
\item \code{4} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/778/vshow}{GSS Data Explorer page for \code{feserve}}.
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\keyword{variable}
