% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{fethink}
\alias{fethink}
\title{How oft r & frnds think abt women's rgts}
\source{
General Social Survey https://gss.norc.org
}
\description{
fethink
}
\details{
Question B. How often would you say that you and your friends think about women's rights?  Would you say that you and your friends think about women's rights very often, fairly often, occasionally, or almost never?
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} fairly often
\item \verb{[3]} occasionally
\item \verb{[4]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/597/vshow}{GSS Data Explorer page for \code{fethink}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab fairly often \tab occasionally \tab almost never \tab don't know \tab no answer \tab Total \cr
   1983 \tab 804 \tab 95 \tab 147 \tab 341 \tab 201 \tab 6 \tab 5 \tab 1599 \cr
   Total \tab 804 \tab 95 \tab 147 \tab 341 \tab 201 \tab 6 \tab 5 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fewrite}}
}
\concept{Core}
\keyword{variable}
