% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{fewknokd}
\alias{fewknokd}
\title{Women should work outside when no kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
fewknokd
}
\details{
Question Do you think that women should work outside the home full-time, part-time or not at all, when a couple has not yet had a child?
}
\section{Values}{

\itemize{
\item \code{1} work full-time
\item \code{2} work part-time
\item \code{3} stay at home
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4454/vshow}{GSS Data Explorer page for \code{fewknokd}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab work full-time \tab work part-time \tab stay at home \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1594 \tab 841 \tab 97 \tab 22 \tab 206 \tab 5 \tab 2765 \cr
   Total \tab 1594 \tab 841 \tab 97 \tab 22 \tab 206 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
