% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{fework}
\alias{fework}
\title{Should women work}
\source{
General Social Survey https://gss.norc.org
}
\description{
fework
}
\details{
Question Do you approve or disapprove of a married woman earning money in business or industry if she has a husband capable of supporting her?
}
\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/589/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab approve \tab disapprove \tab don't know \tab no answer \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 1031 \tab 546 \tab 34 \tab 2 \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab 1504 \tab - \tab 1504 \cr
   1974 \tab 1002 \tab 447 \tab 29 \tab 6 \tab - \tab - \tab 1484 \cr
   1975 \tab 1039 \tab 423 \tab 23 \tab 5 \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab - \tab - \tab 1499 \tab - \tab 1499 \cr
   1977 \tab 999 \tab 507 \tab 22 \tab 2 \tab - \tab - \tab 1530 \cr
   1978 \tab 1107 \tab 402 \tab 21 \tab 2 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab 1468 \tab - \tab 1468 \cr
   1982 \tab 1360 \tab 464 \tab 35 \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 1196 \tab 365 \tab 32 \tab 6 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab 1473 \tab - \tab 1473 \cr
   1985 \tab 1286 \tab 202 \tab 36 \tab 10 \tab - \tab - \tab 1534 \cr
   1986 \tab 1123 \tab 319 \tab 23 \tab 5 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab 1819 \tab - \tab 1819 \cr
   1988 \tab 773 \tab 188 \tab 20 \tab 7 \tab 493 \tab - \tab 1481 \cr
   1989 \tab 774 \tab 208 \tab 18 \tab 6 \tab 531 \tab - \tab 1537 \cr
   1990 \tab 739 \tab 163 \tab 19 \tab 7 \tab 444 \tab - \tab 1372 \cr
   1991 \tab 801 \tab 196 \tab 23 \tab 4 \tab 493 \tab - \tab 1517 \cr
   1993 \tab 849 \tab 207 \tab 22 \tab 2 \tab 526 \tab - \tab 1606 \cr
   1994 \tab 1569 \tab 364 \tab 34 \tab 10 \tab 1015 \tab - \tab 2992 \cr
   1996 \tab 1595 \tab 320 \tab 37 \tab 8 \tab 944 \tab - \tab 2904 \cr
   1998 \tab 1510 \tab 327 \tab 27 \tab 7 \tab 961 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 18753 \tab 5648 \tab 455 \tab 90 \tab 43900 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{kidssol}}

Other Women: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}}
}
\concept{Core}
\concept{Women}
\keyword{variable}
