% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{feworkif}
\alias{feworkif}
\title{Married women work if jobs limited}
\source{
General Social Survey https://gss.norc.org
}
\description{
feworkif
}
\details{
Question If there is a limited number of jobs, do you approve or disapprove of a married woman holding a job in business or industry when her husband is able to support her?
}
\section{Values}{

\itemize{
\item \verb{[1]} approve
\item \verb{[2]} disapprove
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/718/vshow}{GSS Data Explorer page for \code{feworkif}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab approve \tab disapprove \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 543 \tab 953 \tab 30 \tab 4 \tab 1530 \cr
   Total \tab 0 \tab 543 \tab 953 \tab 30 \tab 4 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{racsubgv}},
\code{\link{racsubs}}
}
\concept{Core}
\keyword{variable}
