% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{fidgety}
\alias{fidgety}
\title{In past 30 days feel restless or fidgety}
\source{
General Social Survey https://gss.norc.org
}
\description{
fidgety
}
\details{
Question In the past 30 days, about how often did you feel:
C. Restless or fidgety.
}
\section{Values}{

\itemize{
\item \verb{[1]} all of the time
\item \verb{[2]} most of the time
\item \verb{[3]} some of the time
\item \verb{[4]} little of time
\item \verb{[5]} none of the time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2116/vshow}{GSS Data Explorer page for \code{fidgety}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all of the time \tab most of the time \tab some of the time \tab little of time \tab none of the time \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 52 \tab 132 \tab 412 \tab 431 \tab 397 \tab 5 \tab 16 \tab 2832 \cr
   Total \tab 1387 \tab 52 \tab 132 \tab 412 \tab 431 \tab 397 \tab 5 \tab 16 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{beausprt}},
\code{\link{effort}},
\code{\link{godlove}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{uniongod}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
