% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{fightair}
\alias{fightair}
\title{Women as jet fighter pilots}
\source{
General Social Survey https://gss.norc.org
}
\description{
fightair
}
\details{
Question I'm going to read you a list of jobs that people might have in the armed forces. Please tell me whether you think a woman should or should not be assigned to each job, assuming she is trained to do it.
A. A jet fighter pilot ...
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/789/vshow}{GSS Data Explorer page for \code{fightair}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 1085 \tab 712 \tab 54 \tab 9 \tab 1860 \cr
   Total \tab 0 \tab 1085 \tab 712 \tab 54 \tab 9 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
