% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{fileversion}
\alias{fileversion}
\title{Fileversion}
\source{
General Social Survey https://gss.norc.org
}
\description{
fileversion
}
\details{
Question fileversion
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org}{GSS Data Explorer page for \code{fileversion}}.

Counts by year:\tabular{lll}{
   year \tab 7222.21 \tab Total \cr
   Total \tab 0 \tab 0 \cr
}
}

\keyword{variable}
