% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{fin30}
\alias{fin30}
\title{R use financial information site in past 30 day}
\source{
General Social Survey https://gss.norc.org
}
\description{
fin30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
A. Financial information?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} 3-5 times
\item \code{4} more than 5 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2394/vshow}{GSS Data Explorer page for \code{fin30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 303 \tab 158 \tab 63 \tab 145 \tab 73 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 473 \tab 235 \tab 111 \tab 170 \tab 94 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 361 \tab 153 \tab 54 \tab 108 \tab 13 \tab 1 \tab 2812 \cr
   Total \tab 5878 \tab 1137 \tab 546 \tab 228 \tab 423 \tab 180 \tab 2 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{bookmark}},
\code{\link{catdrcty}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{govt30}},
\code{\link{hyperlnk}},
\code{\link{news30}},
\code{\link{schl30}},
\code{\link{work30}}
}
\concept{Information Society}
\keyword{variable}
