% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_563.R
\name{finan1}
\alias{finan1}
\title{Going bankrupt}
\source{
General Social Survey https://gss.norc.org
}
\description{
finan1
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
C. Did any of the following financial matters happen to you during the last year
\enumerate{
\item Went bankrupt (declared personal bankruptcy).
}
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5160/vshow}{GSS Data Explorer page for \code{finan1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab 500 \tab 13 \tab 1001 \tab 3 \tab - \tab 1517 \cr
   2004 \tab 1472 \tab 17 \tab 1315 \tab 8 \tab - \tab 2812 \cr
   Total \tab 1972 \tab 30 \tab 2316 \tab 11 \tab 0 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{extra1}},
\code{\link{extra2}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famper6}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth12}},
\code{\link{hlth13}},
\code{\link{hlth14}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}},
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}},
\code{\link{oth5}},
\code{\link{prob1}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{work1}},
\code{\link{work10}},
\code{\link{work2}},
\code{\link{work3}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}

Other Income: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{earndes}},
\code{\link{earnmore}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{fairearn}},
\code{\link{famfinan}},
\code{\link{finalter}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{finrela}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{impfinan}},
\code{\link{incdef}},
\code{\link{inclstjb}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{justpay}},
\code{\link{minfood}},
\code{\link{minfour}},
\code{\link{mininc}},
\code{\link{minthree}},
\code{\link{othpay}},
\code{\link{owninc}},
\code{\link{pastpay}},
\code{\link{paychild}},
\code{\link{paychnge}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{payfam1}},
\code{\link{payhard}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{povline}},
\code{\link{realinc}},
\code{\link{realrinc}},
\code{\link{rincblls}},
\code{\link{rincom06}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{salsatfn}},
\code{\link{satfin}},
\code{\link{welfare1}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}}

Other Trauma: 
\code{\link{clsclsex}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}},
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}},
\code{\link{oth5}},
\code{\link{outofbiz}},
\code{\link{ownclerg}},
\code{\link{prob1}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unemp5}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{work1}},
\code{\link{work10}},
\code{\link{work2}},
\code{\link{work3}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}
}
\concept{Income}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
