% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_564.R
\name{finan4}
\alias{finan4}
\title{Being pressured to pay bills}
\source{
General Social Survey https://gss.norc.org
}
\description{
finan4
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
C. Did any of the following financial matters happen to you during the last year
4. Pressured to pay bills by stores, creditors, or bill collectors.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5163/vshow}{GSS Data Explorer page for \code{finan4}}.
}

\seealso{
Other Negative Life Events: 
\code{\link{famper1}},
\code{\link{finan5}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}}

Other Income: 
\code{\link{finan5}}

Other Trauma: 
\code{\link{famper1}},
\code{\link{finan5}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}}
}
\concept{Income}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
