% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{fininc1}
\alias{fininc1}
\title{Importance to r to be told about incentives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fininc1
}
\details{
Question Please assume that your health plan has a financial incentive for your doctor to limit the use of expensive tests.
OR
Please assume that you have a health plan and the plan has a financial incentive for your doctor to limit the use of expensive tests
How important is it for you to be told about these financial incentives?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} somewhat important
\item \verb{[3]} neither important nor unimportant
\item \verb{[4]} somewhat unimportant
\item \verb{[5]} very unimportant
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2748/vshow}{GSS Data Explorer page for \code{fininc1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab somewhat important \tab neither important nor unimportant \tab somewhat unimportant \tab very unimportant \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 2259 \tab 331 \tab 87 \tab 25 \tab 29 \tab 27 \tab 7 \tab 2765 \cr
   Total \tab 0 \tab 2259 \tab 331 \tab 87 \tab 25 \tab 29 \tab 27 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
