% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{fininc2}
\alias{fininc2}
\title{Best time to be told about incentives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fininc2
}
\details{
Question Please assume that your health plan has a financial incentive for your doctor to limit the use of expensive tests.
OR
Please assume that you have a health plan and the plan has a financial incentive for your doctor to limit the use of expensive tests
When do you think is the best time to be told about these financial incentives?
}
\section{Values}{

\itemize{
\item \code{1} when you enroll in a health plan
\item \code{2} when you first see your doctor
\item \code{3} when you need a test or specialist
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2749/vshow}{GSS Data Explorer page for \code{fininc2}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
