% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{fininc3}
\alias{fininc3}
\title{Best person to tell about incentives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fininc3
}
\details{
Question Please assume that your health plan has a financial incentive for your doctor to limit the use of expensive tests.
OR
Please assume that you have a health plan and the plan has a financial incentive for your doctor to limit the use of expensive tests
Who is the best person to tell you about these financial incentives?
}
\section{Values}{

\itemize{
\item \verb{[1]} your doctor
\item \verb{[2]} a representative of your health plan
\item \verb{[3]} both your doctor and a representative of health plan
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2750/vshow}{GSS Data Explorer page for \code{fininc3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab your doctor \tab a representative of your health plan \tab both your doctor and a representative of health plan \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 465 \tab 1217 \tab 1034 \tab 42 \tab 7 \tab 2765 \cr
   Total \tab 0 \tab 465 \tab 1217 \tab 1034 \tab 42 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
