% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{fininc4}
\alias{fininc4}
\title{R likely to ask doc about incentives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fininc4
}
\details{
Question Please assume that your health plan has a financial incentive for your doctor to limit the use of expensive tests.
OR
Please assume that you have a health plan and the plan has a financial incentive for your doctor to limit the use of expensive tests
How likely would you be to ask your regular doctor whether he or she has financial incentives that limit the use of expensive tests?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} neither likely nor unlikely
\item \code{4} somewhat unlikely
\item \code{5} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2751/vshow}{GSS Data Explorer page for \code{fininc4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab neither likely nor unlikely \tab somewhat unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 1030 \tab 682 \tab 286 \tab 365 \tab 350 \tab 42 \tab 10 \tab 2765 \cr
   Total \tab 0 \tab 1030 \tab 682 \tab 286 \tab 365 \tab 350 \tab 42 \tab 10 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
