% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_390.R
\name{fireftng}
\alias{fireftng}
\title{How scientific is firefighting}
\source{
General Social Survey https://gss.norc.org
}
\description{
fireftng
}
\details{
Question C. Firefighting. (Is firefighting...?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very scientific
\item \verb{[2]} pretty scientific
\item \verb{[3]} not too scientific
\item \verb{[4]} not scientific at all
\item \verb{[5]} haven't heard of it (volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3539/vshow}{GSS Data Explorer page for \code{fireftng}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab haven't heard of it (volunteered) \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1469 \tab 88 \tab 193 \tab 143 \tab 62 \tab 1 \tab 15 \tab 3 \tab 1974 \cr
   Total \tab 1469 \tab 88 \tab 193 \tab 143 \tab 62 \tab 1 \tab 15 \tab 3 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{architct}},
\code{\link{cmprgmng}},
\code{\link{engnring}},
\code{\link{finlcoun}},
\code{\link{lawenfrc}},
\code{\link{marrcoun}},
\code{\link{medtreat}},
\code{\link{slsmnshp}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
