% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{firstjob}
\alias{firstjob}
\title{Type of first job held by r}
\source{
General Social Survey https://gss.norc.org
}
\description{
firstjob
}
\details{
Question And how about your first job -- the first job you had after you finished full-time education?
}
\section{Values}{

\itemize{
\item \verb{[1]} prof or tech
\item \verb{[2]} higher admin
\item \verb{[3]} clerical
\item \verb{[4]} sales
\item \verb{[5]} service
\item \verb{[6]} skilled worker
\item \verb{[7]} semi-skilled worker
\item \verb{[8]} unskilled worker
\item \verb{[9]} farm
\item \verb{[10]} never had job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4316/vshow}{GSS Data Explorer page for \code{firstjob}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab prof or tech \tab higher admin \tab clerical \tab sales \tab service \tab skilled worker \tab semi-skilled worker \tab unskilled worker \tab farm \tab never had job \tab no answer \tab Total \cr
   1987 \tab 255 \tab 229 \tab 9 \tab 309 \tab 94 \tab 134 \tab 116 \tab 120 \tab 319 \tab 73 \tab 55 \tab 106 \tab 1819 \cr
   Total \tab 255 \tab 229 \tab 9 \tab 309 \tab 94 \tab 134 \tab 116 \tab 120 \tab 319 \tab 73 \tab 55 \tab 106 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
