% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{firstkidsex}
\alias{firstkidsex}
\title{Sex of first child}
\source{
General Social Survey https://gss.norc.org
}
\description{
firstkidsex
}
\details{
Question Was your first biological child's sex recorded as male or female at birth?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7958/vshow}{GSS Data Explorer page for \code{firstkidsex}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab skipped on web \tab Total \cr
   2022 \tab 2844 \tab 316 \tab 382 \tab 2 \tab 3544 \cr
   Total \tab 2844 \tab 316 \tab 382 \tab 2 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{femself}},
\code{\link{mascself}},
\code{\link{nonbinkids}}
}
\concept{GSS Next}
\keyword{variable}
