% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{firstsei}
\alias{firstsei}
\title{R's socioeconomic index of 1st occupation (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
firstsei
}
\details{
Question Socioeconomic index of respondent’s first occupation.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} i don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5480/vshow}{GSS Data Explorer page for \code{firstsei}}.
}

\seealso{
Other Occupation: 
\code{\link{coisco08}},
\code{\link{masei}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{spsei}}
}
\concept{Occupation}
\keyword{variable}
