% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{firstyou}
\alias{firstyou}
\title{Should take care of oneself and ones family first}
\source{
General Social Survey https://gss.norc.org
}
\description{
firstyou
}
\details{
Question Please select a response to indicate how much you agree or disagree with each of the following statements.
B. You should take care of yourself and your family first, before helping other people
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4230/vshow}{GSS Data Explorer page for \code{firstyou}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{helpjob}},
\code{\link{kidpars}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{talkedto}}
}
\concept{Altruism}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
