% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{fishexp1}
\alias{fishexp1}
\title{What is scientist trying to find out from this experiment}
\source{
General Social Survey https://gss.norc.org
}
\description{
fishexp1
}
\details{
Question Please look at Card I-5. What is the scientist trying to find out from this experiment?
}
\section{Values}{

\itemize{
\item \code{1} if the number of fish in the fish bowl affects the behavior of the fish
\item \code{2} if the temperature of the fish bowl affects the behavior of the fish
\item \code{3} if the temperature and the amount of light affect the behavior of the fish
\item \code{4} if the number of fish, the temperature, and the amount of light affect the behavior of the fish
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3485/vshow}{GSS Data Explorer page for \code{fishexp1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab if the number of fish in the fish bowl affects the behavior of the fish \tab if the temperature of the fish bowl affects the behavior of the fish \tab if the temperature and the amount of light affect the behavior of the fish \tab if the number of fish, the temperature, and the amount of light affect the behavior of the fish \tab don't know \tab no answer \tab Total \cr
   2008 \tab 709 \tab 520 \tab 85 \tab 71 \tab 455 \tab 155 \tab 28 \tab 2023 \cr
   Total \tab 709 \tab 520 \tab 85 \tab 71 \tab 455 \tab 155 \tab 28 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
