% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{fishexp2}
\alias{fishexp2}
\title{Why did you choose that answer}
\source{
General Social Survey https://gss.norc.org
}
\description{
fishexp2
}
\details{
Question Why did you choose that answer?
}
\section{Values}{

\itemize{
\item \verb{[1]} because i already know what affects the behavior of fish
\item \verb{[2]} because that is what is allowed to change in this experiment
\item \verb{[3]} because that is what stays the same in this experiment
\item \verb{[4]} because that is what the scientist decided to include in this experiment
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3486/vshow}{GSS Data Explorer page for \code{fishexp2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab because i already know what affects the behavior of fish \tab because that is what is allowed to change in this experiment \tab because that is what stays the same in this experiment \tab because that is what the scientist decided to include in this experiment \tab don't know \tab no answer \tab Total \cr
   2008 \tab 709 \tab 131 \tab 489 \tab 151 \tab 278 \tab 236 \tab 29 \tab 2023 \cr
   Total \tab 709 \tab 131 \tab 489 \tab 151 \tab 278 \tab 236 \tab 29 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
