% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{fixerd}
\alias{fixerd}
\title{General reputation in the community}
\source{
General Social Survey https://gss.norc.org
}
\description{
fixerd
}
\details{
Question Think about the most recent time you've paid for home maintenance service -- e.g. a roofer, a painter, a plumber or something like that. What made you choose the person or business that did the work? (Choose all that apply.)
D. General reputation in the community.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1843/vshow}{GSS Data Explorer page for \code{fixerd}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2353 \tab 123 \tab 292 \tab 1 \tab 135 \tab 2904 \cr
   Total \tab 2353 \tab 123 \tab 292 \tab 1 \tab 135 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{fixer1}},
\code{\link{fixerc}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{relfixer}},
\code{\link{satfixes}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
