% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{fixmistk}
\alias{fixmistk}
\title{How likely govt admnstrtor's be corrected when making}
\source{
General Social Survey https://gss.norc.org
}
\description{
fixmistk
}
\details{
Question When the government administrators make serious mistakes in America how likely is it that they will be corrected?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4958/vshow}{GSS Data Explorer page for \code{fixmistk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 126 \tab 591 \tab 590 \tab 142 \tab 17 \tab 6 \tab 2812 \cr
   Total \tab 1340 \tab 126 \tab 591 \tab 590 \tab 142 \tab 17 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
