% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{flexhrs}
\alias{flexhrs}
\title{Importance of flexible hrs. in a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
flexhrs
}
\details{
Question On the following list there are various aspects of jobs. Please circle one number to show how important you personally consider it is in a job:
I. A job with flexible working hours?
}
\section{Values}{

\itemize{
\item \code{1} very impt
\item \code{2} impt
\item \code{3} neither
\item \code{4} not impt
\item \code{5} not impt at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4515/vshow}{GSS Data Explorer page for \code{flexhrs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very impt \tab impt \tab neither \tab not impt \tab not impt at all \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 218 \tab 570 \tab 415 \tab 171 \tab 31 \tab 20 \tab 28 \tab 1537 \cr
   1998 \tab 1609 \tab 197 \tab 441 \tab 385 \tab 118 \tab 16 \tab 27 \tab 39 \tab 2832 \cr
   2006 \tab 2992 \tab 264 \tab 553 \tab 395 \tab 243 \tab 52 \tab 9 \tab 2 \tab 4510 \cr
   Total \tab 4685 \tab 679 \tab 1564 \tab 1195 \tab 532 \tab 99 \tab 56 \tab 69 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
