% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{flexhrs1}
\alias{flexhrs1}
\title{Importance of flexible hrs. at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
flexhrs1
}
\details{
Question (HAND CARD D2)
(For each of the following, please tell me how important you personally think it is in a job. How important is...)
...a job that allows someone to decide their times or days of work?
(Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} neither important nor unimportant
\item \code{4} not important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6321/vshow}{GSS Data Explorer page for \code{flexhrs1}}.
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
