% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{fluvax}
\alias{fluvax}
\title{Flu vaccine in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
fluvax
}
\details{
Question Did you receive a seasonal influenza vaccine ("flu shot") in the past 12 months?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7951/vshow}{GSS Data Explorer page for \code{fluvax}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 694 \tab 535 \tab 6 \tab 3544 \cr
   Total \tab 2309 \tab 694 \tab 535 \tab 6 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{malekids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
