% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{fndjob4}
\alias{fndjob4}
\title{Heard job from acquaintances}
\source{
General Social Survey https://gss.norc.org
}
\description{
fndjob4
}
\details{
Question There are many ways people hear about jobs -- such as from ads, employment agencies, or from other people. Could you tell me how you found out about work at your present employer (IF SELF-EMPLOYED:  current organization)?  Please tell me all the ways that you found out about this job.
D. From an acquaintance.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1279/vshow}{GSS Data Explorer page for \code{fndjob4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 605 \tab 102 \tab 686 \tab 124 \tab 1517 \cr
   Total \tab 605 \tab 102 \tab 686 \tab 124 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{numunemp}},
\code{\link{orgfin}},
\code{\link{totunemp}},
\code{\link{wrkyears}}
}
\concept{Work Organizations}
\keyword{variable}
