% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_137.R
\name{fndjob7}
\alias{fndjob7}
\title{Heard job from recruiter for this employer}
\source{
General Social Survey https://gss.norc.org
}
\description{
fndjob7
}
\details{
Question There are many ways people hear about jobs -- such as from ads, employment agencies, or from other people. Could you tell me how you found out about work at your present employer (IF SELF-EMPLOYED:  current organization)?  Please tell me all the ways that you found out about this job.
G. From a recruiter for this employer.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1282/vshow}{GSS Data Explorer page for \code{fndjob7}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 605 \tab 38 \tab 742 \tab 132 \tab 1517 \cr
   Total \tab 605 \tab 38 \tab 742 \tab 132 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{drugtest}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{intltest}},
\code{\link{ltrsref}},
\code{\link{physical}},
\code{\link{skiltest}}
}
\concept{Work Organizations}
\keyword{variable}
