% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_137.R
\name{fndjob8}
\alias{fndjob8}
\title{Heard job just came here and asked for work}
\source{
General Social Survey https://gss.norc.org
}
\description{
fndjob8
}
\details{
Question There are many ways people hear about jobs -- such as from ads, employment agencies, or from other people. Could you tell me how you found out about work at your present employer (IF SELF-EMPLOYED:  current organization)?  Please tell me all the ways that you found out about this job.
H. I just came here and asked for work.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1283/vshow}{GSS Data Explorer page for \code{fndjob8}}.
}

\seealso{
Other Work Organizations: 
\code{\link{drugtest}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob9}},
\code{\link{intltest}},
\code{\link{ltrsref}},
\code{\link{physical}},
\code{\link{skiltest}}

Other Employment: 
\code{\link{drugtest}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob9}},
\code{\link{intltest}},
\code{\link{ltrsref}},
\code{\link{physical}},
\code{\link{skiltest}}
}
\concept{Employment}
\concept{Work Organizations}
\keyword{variable}
