% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_137.R
\name{fndjob9}
\alias{fndjob9}
\title{Heard job other (specify)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fndjob9
}
\details{
Question There are many ways people hear about jobs -- such as from ads, employment agencies, or from other people. Could you tell me how you found out about work at your present employer (IF SELF-EMPLOYED:  current organization)?  Please tell me all the ways that you found out about this job.
I. Other (SPECIFY).
}
\section{Values}{

\itemize{
\item \verb{[1]} union
\item \verb{[2]} school
\item \verb{[3]} self-employed,created job
\item \verb{[4]} announcement,newspaper
\item \verb{[7]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1284/vshow}{GSS Data Explorer page for \code{fndjob9}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab union \tab school \tab self-employed,created job \tab announcement,newspaper \tab other \tab no answer \tab Total \cr
   1991 \tab 605 \tab 10 \tab 1 \tab 22 \tab 7 \tab 21 \tab 851 \tab 1517 \cr
   Total \tab 605 \tab 10 \tab 1 \tab 22 \tab 7 \tab 21 \tab 851 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{drugtest}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{intltest}},
\code{\link{ltrsref}},
\code{\link{physical}},
\code{\link{skiltest}}
}
\concept{Work Organizations}
\keyword{variable}
