% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{follow}
\alias{follow}
\title{Importance of following church teachings}
\source{
General Social Survey https://gss.norc.org
}
\description{
follow
}
\details{
Question People have many differing views about what makes a person a good Christian or Jew. Please tell me how important is each of the following to you. Once again, please indicate where you would place your feelings on a scale from 1 to 5. How important is it …
C. To follow faithfully the teachings of their church or synagogue.
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[5]} not very important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1091/vshow}{GSS Data Explorer page for \code{follow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab 2 \tab 3 \tab 4 \tab not very important \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 465 \tab 331 \tab 353 \tab 155 \tab 152 \tab 12 \tab 13 \tab 1481 \cr
   1998 \tab 1387 \tab 457 \tab 295 \tab 336 \tab 127 \tab 190 \tab 28 \tab 12 \tab 2832 \cr
   Total \tab 1387 \tab 922 \tab 626 \tab 689 \tab 282 \tab 342 \tab 40 \tab 25 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
