% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_682.R
\name{foodalc}
\alias{foodalc}
\title{Would x engage in excessive and uncontrolled food or alcohol consumption}
\source{
General Social Survey https://gss.norc.org
}
\description{
foodalc
}
\details{
Question How likely in your opinion is it (VIGNAME) would engage in excessive and uncontrolled food or alcohol consumption?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8135/vshow}{GSS Data Explorer page for \code{foodalc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 177 \tab 695 \tab 492 \tab 179 \tab 60 \tab 1664 \tab 6 \tab 36 \tab 3309 \cr
   Total \tab 0 \tab 177 \tab 695 \tab 492 \tab 179 \tab 60 \tab 1664 \tab 6 \tab 36 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{bullpeer}},
\code{\link{cutself}},
\code{\link{fighpeer}},
\code{\link{negself}},
\code{\link{physill1}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{schshoot}},
\code{\link{takelife}}
}
\concept{Mental Health}
\keyword{variable}
