% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_330.R
\name{forcedoc}
\alias{forcedoc}
\title{Should child be forced by law to be treated by doc}
\source{
General Social Survey https://gss.norc.org
}
\description{
forcedoc
}
\details{
Question A. Do you think that parents of children like (John/Mary) should be forced by law to have (John/Mary) treated
at a clinic or by a doctor?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2949/vshow}{GSS Data Explorer page for \code{forcedoc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 338 \tab 1001 \tab 42 \tab 12 \tab 2765 \cr
   Total \tab 1372 \tab 338 \tab 1001 \tab 42 \tab 12 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcoumed}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{grwingup}},
\code{\link{mntlill}}
}
\concept{Mental Health II}
\keyword{variable}
