% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{forgive1}
\alias{forgive1}
\title{Have forgiven self for mistakes}
\source{
General Social Survey https://gss.norc.org
}
\description{
forgive1
}
\details{
Question Because of your religious or spiritual beliefs have you always, almost always, often, seldom, or never felt the following ways:
A. I have forgiven myself for things that I have done wrong.
}
\section{Values}{

\itemize{
\item \code{1} alwys/almst alwy
\item \code{2} often
\item \code{3} seldom
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2105/vshow}{GSS Data Explorer page for \code{forgive1}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{relcmfrt}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
