% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{forland}
\alias{forland}
\title{Foreigner's should not be allowed to buy land}
\source{
General Social Survey https://gss.norc.org
}
\description{
forland
}
\details{
Question How much do you agree or disagree with the following statements?
E. Foreigners should not be allowed to buy land in America
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4854/vshow}{GSS Data Explorer page for \code{forland}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 147 \tab 278 \tab 350 \tab 401 \tab 124 \tab 45 \tab 22 \tab 2904 \cr
   2004 \tab 1596 \tab 139 \tab 216 \tab 277 \tab 446 \tab 110 \tab 27 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 103 \tab 186 \tab 251 \tab 535 \tab 144 \tab 51 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 389 \tab 680 \tab 878 \tab 1382 \tab 378 \tab 123 \tab 27 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
