% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{forlang}
\alias{forlang}
\title{American schools should teach foreign languages}
\source{
General Social Survey https://gss.norc.org
}
\description{
forlang
}
\details{
Question How much do you agree or disagree with the following statements?
C. American schools should make much more effort to teach foreign languages properly
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4852/vshow}{GSS Data Explorer page for \code{forlang}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 236 \tab 600 \tab 347 \tab 119 \tab 15 \tab 29 \tab 21 \tab 2904 \cr
   Total \tab 1537 \tab 236 \tab 600 \tab 347 \tab 119 \tab 15 \tab 29 \tab 21 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
