% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_254.R
\name{forlang1}
\alias{forlang1}
\title{Learning foreign language important math and s}
\source{
General Social Survey https://gss.norc.org
}
\description{
forlang1
}
\details{
Question Now please tell us whether you strongly agree, agree, disagree or strongly disagree with each of these statements:
D. Learning a foreign language is as valuable as learning math and science in school.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2298/vshow}{GSS Data Explorer page for \code{forlang1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 278 \tab 583 \tab 422 \tab 70 \tab 36 \tab 9 \tab 2817 \cr
   Total \tab 1419 \tab 278 \tab 583 \tab 422 \tab 70 \tab 36 \tab 9 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{twolang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
