% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{formldys}
\alias{formldys}
\title{The number of days in formal training}
\source{
General Social Survey https://gss.norc.org
}
\description{
formldys
}
\details{
Question D. In total, did the total number of days add up to...
}
\section{Values}{

\itemize{
\item \verb{[1]} a month or less
\item \verb{[2]} 1-3 months
\item \verb{[3]} 3-6 months
\item \verb{[4]} 6 months-1 yr
\item \verb{[5]} over a year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2152/vshow}{GSS Data Explorer page for \code{formldys}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a month or less \tab 1-3 months \tab 3-6 months \tab 6 months-1 yr \tab over a year \tab no answer \tab Total \cr
   1998 \tab 2377 \tab 163 \tab 74 \tab 58 \tab 47 \tab 74 \tab 39 \tab 2832 \cr
   Total \tab 2377 \tab 163 \tab 74 \tab 58 \tab 47 \tab 74 \tab 39 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
