% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{formlfin}
\alias{formlfin}
\title{Have you finished all the courses}
\source{
General Social Survey https://gss.norc.org
}
\description{
formlfin
}
\details{
Question C. Have you finished all the required courses?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} never ends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2151/vshow}{GSS Data Explorer page for \code{formlfin}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab never ends \tab no answer \tab Total \cr
   1998 \tab 2325 \tab 398 \tab 21 \tab 52 \tab 36 \tab 2832 \cr
   Total \tab 2325 \tab 398 \tab 21 \tab 52 \tab 36 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
