% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_152.R
\name{frdcreat}
\alias{frdcreat}
\title{Desir qualities of friend: creative}
\source{
General Social Survey https://gss.norc.org
}
\description{
frdcreat
}
\details{
Question I'm going to read seven qualities one might look for in a personal friend. All of the qualities may be desirable ones for a personal friend, but I'm interested in those that are most important to you. (HAND CARD CB TO RESPONDENT.)
As I read each one, could you tell me whether it is extremely important, very important, fairly important
ot too important, or not at all important.
A. Creative.
}
\section{Values}{

\itemize{
\item \code{1} extremely important
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not too important
\item \code{5} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1431/vshow}{GSS Data Explorer page for \code{frdcreat}}.
}

\seealso{
Other Culture: 
\code{\link{excelart}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{hosthome}},
\code{\link{impfinan}}
}
\concept{Culture}
\keyword{variable}
