% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_152.R
\name{frdcultr}
\alias{frdcultr}
\title{Desir qualities of friend: cultured}
\source{
General Social Survey https://gss.norc.org
}
\description{
frdcultr
}
\details{
Question I'm going to read seven qualities one might look for in a personal friend. All of the qualities may be desirable ones for a personal friend, but I'm interested in those that are most important to you. (HAND CARD CB TO RESPONDENT.)
As I read each one, could you tell me whether it is extremely important, very important, fairly important
ot too important, or not at all important.
B. Cultured.
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely important
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not too important
\item \verb{[5]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1432/vshow}{GSS Data Explorer page for \code{frdcultr}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab extremely important \tab very important \tab fairly important \tab not too important \tab not at all important \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1993 \tab - \tab 52 \tab 275 \tab 596 \tab 494 \tab 148 \tab 26 \tab 15 \tab - \tab 1606 \cr
   Total \tab 0 \tab 52 \tab 275 \tab 596 \tab 494 \tab 148 \tab 26 \tab 15 \tab 0 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{bigband}},
\code{\link{blues}},
\code{\link{blugrass}},
\code{\link{camping}},
\code{\link{classicl}},
\code{\link{classics}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{conrock}},
\code{\link{country}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{english}},
\code{\link{excelart}},
\code{\link{folk}},
\code{\link{frdcreat}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{gospel}},
\code{\link{grtbooks}},
\code{\link{hosthome}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{huntfish}},
\code{\link{hvymetal}},
\code{\link{impcultr}},
\code{\link{impfinan}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{jazz}},
\code{\link{judgeart}},
\code{\link{latin}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{makeart}},
\code{\link{modpaint}},
\code{\link{moodeasy}},
\code{\link{mostimp1}},
\code{\link{mostimp2}},
\code{\link{mostimp3}},
\code{\link{mostimp4}},
\code{\link{musicals}},
\code{\link{newage}},
\code{\link{obtohelp}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{opoutcme}},
\code{\link{pclit}},
\code{\link{perform}},
\code{\link{plymusic}},
\code{\link{popmusic}},
\code{\link{rap}},
\code{\link{readfict}},
\code{\link{reggae}},
\code{\link{richpoor}},
\code{\link{seemovie}},
\code{\link{selfirst}},
\code{\link{standup}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}},
\code{\link{united}},
\code{\link{usevcr}},
\code{\link{visitart}}

Other Friends: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{bwfriend}},
\code{\link{clsclsex}},
\code{\link{cowrkfrd}},
\code{\link{frdcreat}},
\code{\link{frddynam}},
\code{\link{frdenjoy}},
\code{\link{frdfun}},
\code{\link{frdhelps}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdknows}},
\code{\link{frdresp}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{friends}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndknow}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{makefrnd}},
\code{\link{needyfrd}},
\code{\link{neifrd}},
\code{\link{numbwfrd}},
\code{\link{numfrend}},
\code{\link{othfrd}},
\code{\link{rightpeo}}
}
\concept{Culture}
\concept{Friends}
\keyword{variable}
