% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_152.R
\name{frddynam}
\alias{frddynam}
\title{Desir qualities of friend: dynamic}
\source{
General Social Survey https://gss.norc.org
}
\description{
frddynam
}
\details{
Question I'm going to read seven qualities one might look for in a personal friend. All of the qualities may be desirable ones for a personal friend, but I'm interested in those that are most important to you. (HAND CARD CB TO RESPONDENT.)
As I read each one, could you tell me whether it is extremely important, very important, fairly important
ot too important, or not at all important.
C. Dynamic.
}
\section{Values}{

\itemize{
\item \code{1} extremely important
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not too important
\item \code{5} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1433/vshow}{GSS Data Explorer page for \code{frddynam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely important \tab very important \tab fairly important \tab not too important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 50 \tab 254 \tab 591 \tab 520 \tab 135 \tab 41 \tab 15 \tab 1606 \cr
   Total \tab 0 \tab 50 \tab 254 \tab 591 \tab 520 \tab 135 \tab 41 \tab 15 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{excelart}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{hosthome}},
\code{\link{impfinan}}
}
\concept{Culture}
\keyword{variable}
