% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{freeexp1}
\alias{freeexp1}
\title{What experiences in life make you feel most free}
\source{
General Social Survey https://gss.norc.org
}
\description{
freeexp1
}
\details{
Question What experiences in your life make you feel most free?  Are there times in your life when you feel especially  free?  Can you give me an example of an experience that makes you feel really free?
A. First distinct experience mentioned
}
\section{Values}{

\itemize{
\item \verb{[1]} moving:driving travel
\item \verb{[2]} leisure:recreation enjoyable activity
\item \verb{[3]} social:engagement with family or institutions
\item \verb{[4]} negative:no obigation, divorced, being alone
\item \verb{[5]} inner:spiritual and religious freedom
\item \verb{[6]} choices:variant of making choices
\item \verb{[7]} citizen:political participation
\item \verb{[8]} security:economic security and independence
\item \verb{[9]} other
\item \verb{[98]} no answer dont know
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2660/vshow}{GSS Data Explorer page for \code{freeexp1}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab moving:driving travel \tab leisure:recreation enjoyable activity \tab social:engagement with family or institutions \tab negative:no obigation, divorced, being alone \tab inner:spiritual and religious freedom \tab choices:variant of making choices \tab citizen:political participation \tab security:economic security and independence \tab other \tab no answer \tab Total \cr
   2000 \tab 1403 \tab 235 \tab 131 \tab 94 \tab 120 \tab 134 \tab 81 \tab 163 \tab 197 \tab 98 \tab 161 \tab 2817 \cr
   Total \tab 1403 \tab 235 \tab 131 \tab 94 \tab 120 \tab 134 \tab 81 \tab 163 \tab 197 \tab 98 \tab 161 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}}
}
\concept{Freedom}
\keyword{variable}
