% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{freeexp2}
\alias{freeexp2}
\title{What experiences in life make you feel most free}
\source{
General Social Survey https://gss.norc.org
}
\description{
freeexp2
}
\details{
Question What experiences in your life make you feel most free?  Are there times in your life when you feel especially  free?  Can you give me an example of an experience that makes you feel really free?
B. Second distinct experience mentioned
}
\section{Values}{

\itemize{
\item \code{1} moving:driving travel
\item \code{2} leisure:recreation enjoyable activity
\item \code{3} social:engagement with family or institutions
\item \code{4} negative:no obigation, divorced, being alone
\item \code{5} inner:spiritual and religious freedom
\item \code{6} choices:variant of making choices
\item \code{7} citizen:political participation
\item \code{8} security:economic security and independence
\item \code{9} other
\item \code{98} no answer dont know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2661/vshow}{GSS Data Explorer page for \code{freeexp2}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab moving:driving travel \tab leisure:recreation enjoyable activity \tab social:engagement with family or institutions \tab negative:no obigation, divorced, being alone \tab inner:spiritual and religious freedom \tab choices:variant of making choices \tab citizen:political participation \tab security:economic security and independence \tab other \tab Total \cr
   2000 \tab 2400 \tab 78 \tab 38 \tab 58 \tab 32 \tab 48 \tab 15 \tab 53 \tab 79 \tab 16 \tab 2817 \cr
   Total \tab 2400 \tab 78 \tab 38 \tab 58 \tab 32 \tab 48 \tab 15 \tab 53 \tab 79 \tab 16 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{freeexp1}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}}
}
\concept{Freedom}
\keyword{variable}
