% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{freeexp3}
\alias{freeexp3}
\title{What experiences in life make you feel most free}
\source{
General Social Survey https://gss.norc.org
}
\description{
freeexp3
}
\details{
Question What experiences in your life make you feel most free?  Are there times in your life when you feel especially  free?  Can you give me an example of an experience that makes you feel really free?
C. Third distinct experience mentioned
}
\section{Values}{

\itemize{
\item \code{1} moving:driving travel
\item \code{2} leisure:recreation enjoyable activity
\item \code{3} social:engagement with family or institutions
\item \code{4} negative:no obigation, divorced, being alone
\item \code{5} inner:spiritual and religious freedom
\item \code{6} choices:variant of making choices
\item \code{7} citizen:political participation
\item \code{8} security:economic security and independence
\item \code{9} other
\item \code{98} no answer dont know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2662/vshow}{GSS Data Explorer page for \code{freeexp3}}.
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{freenow}},
\code{\link{freeprss}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}},
\code{\link{wlthpov}}
}
\concept{Freedom}
\keyword{variable}
