% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{freemns3}
\alias{freemns3}
\title{What does freedom mean to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
freemns3
}
\details{
Question The next questions are about freedom in American today. Freedom means many different things to people. When you think about freedom, what comes to mind? Can you tell me in a couple of sentences what does freedom mean to you?
C. Third distinct subject mentioned
}
\section{Values}{

\itemize{
\item \verb{[1]} want:doing what i want
\item \verb{[2]} autonomy:self-realization
\item \verb{[3]} choices:making decision
\item \verb{[4]} rights:religion and speech
\item \verb{[5]} movement:movement and travel
\item \verb{[6]} negative:slave surbordinate discrimination
\item \verb{[7]} citizen:political participation
\item \verb{[8]} security:economic independence
\item \verb{[9]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2659/vshow}{GSS Data Explorer page for \code{freemns3}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab want:doing what i want \tab autonomy:self-realization \tab choices:making decision \tab rights:religion and speech \tab movement:movement and travel \tab negative:slave surbordinate discrimination \tab citizen:political participation \tab security:economic independence \tab other \tab Total \cr
   2000 \tab 2623 \tab 28 \tab 9 \tab 21 \tab 53 \tab 14 \tab 18 \tab 16 \tab 17 \tab 18 \tab 2817 \cr
   Total \tab 2623 \tab 28 \tab 9 \tab 21 \tab 53 \tab 14 \tab 18 \tab 16 \tab 17 \tab 18 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}}
}
\concept{Freedom}
\keyword{variable}
