% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{freenow}
\alias{freenow}
\title{Americans have more freedom than in past}
\source{
General Social Survey https://gss.norc.org
}
\description{
freenow
}
\details{
Question Do you think Americans today have more freedom, less freedom, or about the same amount of freedom as in the past?
}
\section{Values}{

\itemize{
\item \verb{[1]} more freedom
\item \verb{[2]} about the same amount of freedom
\item \verb{[3]} less freedom
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2665/vshow}{GSS Data Explorer page for \code{freenow}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab more freedom \tab about the same amount of freedom \tab less freedom \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 585 \tab 536 \tab 262 \tab 18 \tab 18 \tab 2817 \cr
   Total \tab 1398 \tab 585 \tab 536 \tab 262 \tab 18 \tab 18 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}}
}
\concept{Freedom}
\keyword{variable}
