% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{freeprss}
\alias{freeprss}
\title{Press should be completely free}
\source{
General Social Survey https://gss.norc.org
}
\description{
freeprss
}
\details{
Question Many people think it is important to have complete freedom of the press and news media in the United States. Others think that the press too often invades the privacy of public figures–like Senators or members of Congress by printing stories that contain personal details about their private lives. Which of the three statements on this Card comes closest to your feelings about balancing freedom of the press and the right to privacy?
READ STATEMENTS 1-3 ONLY. CIRCLE ONLY ONE CODE.
}
\section{Values}{

\itemize{
\item \code{1} there should be complete freedom of the press, even if
\item \code{2} the press should develop a code of ethics to keep it
\item \code{3} the government should keep the press from printing
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2674/vshow}{GSS Data Explorer page for \code{freeprss}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab there should be complete freedom of the press, even if \tab the press should develop a code of ethics to keep it \tab the government should keep the press from printing \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 292 \tab 879 \tab 200 \tab 31 \tab 17 \tab 2817 \cr
   Total \tab 1398 \tab 292 \tab 879 \tab 200 \tab 31 \tab 17 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{wlthpov}}
}
\concept{Freedom}
\keyword{variable}
