% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{freetrde}
\alias{freetrde}
\title{Free trade leads to better products}
\source{
General Social Survey https://gss.norc.org
}
\description{
freetrde
}
\details{
Question How much do you agree or disagree with the following statements? (Please, select one response)
B. Free trade leads to better products becoming available in America.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4884/vshow}{GSS Data Explorer page for \code{freetrde}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 134 \tab 525 \tab 297 \tab 188 \tab 23 \tab 48 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 105 \tab 582 \tab 305 \tab 146 \tab 37 \tab 95 \tab 4 \tab 2538 \cr
   Total \tab 2860 \tab 239 \tab 1107 \tab 602 \tab 334 \tab 60 \tab 143 \tab 5 \tab 5350 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
