% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{fricall}
\alias{fricall}
\title{Contact with closest friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
fricall
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to…
F. And how often do you have any other contact with this friend besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} several x week
\item \code{3} once a week
\item \code{4} once a month
\item \code{5} several x year
\item \code{6} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4168/vshow}{GSS Data Explorer page for \code{fricall}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}

Other Friends: 
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}}

Other Social Networks: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}}
}
\concept{Friends}
\concept{ISSP Social Networks & Support Systems}
\concept{Social Networks}
\keyword{variable}
