% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_100.R
\name{friend3}
\alias{friend3}
\title{Person #3 is r's friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
friend3
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
I3. Third Person is Friend
}
\section{Values}{

\itemize{
\item \code{1} mentioned
\item \code{2} not mentioned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/917/vshow}{GSS Data Explorer page for \code{friend3}}.
}

\seealso{
Other Social Networks: 
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}}

Other Social Networks: 
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}}
}
\concept{Social Networks}
\keyword{variable}
