% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{frineigh}
\alias{frineigh}
\title{Number of close friends are neighbors}
\source{
General Social Survey https://gss.norc.org
}
\description{
frineigh
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to… How many close friends would you say you have?
B. How many of these close friends are your neighbors now?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4164/vshow}{GSS Data Explorer page for \code{frineigh}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 14 \tab 15 \tab 17 \tab 18 \tab 2 \tab 20 \tab 25 \tab 3 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab 9 \tab 96 \tab don't know \tab no answer \tab Total \cr
   1986 \tab 77 \tab 768 \tab 256 \tab 18 \tab 1 \tab 7 \tab 1 \tab 4 \tab 1 \tab 1 \tab 140 \tab 5 \tab 5 \tab 71 \tab 34 \tab 30 \tab 1 \tab 23 \tab 4 \tab 6 \tab 4 \tab 1 \tab 1 \tab 11 \tab 1470 \cr
   Total \tab 77 \tab 768 \tab 256 \tab 18 \tab 1 \tab 7 \tab 1 \tab 4 \tab 1 \tab 1 \tab 140 \tab 5 \tab 5 \tab 71 \tab 34 \tab 30 \tab 1 \tab 23 \tab 4 \tab 6 \tab 4 \tab 1 \tab 1 \tab 11 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
