% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{fringe10}
\alias{fringe10}
\title{Information about child care services in company}
\source{
General Social Survey https://gss.norc.org
}
\description{
fringe10
}
\details{
Question I am going to read a list of fringe benefits that workers sometimes get in addition to their wages. Whether you receive it or not, please tell me whether you are eligible to receive each fringe benefit.
J. Information about child care services in the community.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1362/vshow}{GSS Data Explorer page for \code{fringe10}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 178 \tab 673 \tab 47 \tab 14 \tab 1517 \cr
   Total \tab 605 \tab 178 \tab 673 \tab 47 \tab 14 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{cowrkrel}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{jobvshme}},
\code{\link{schoolng}},
\code{\link{unmanrel}},
\code{\link{unvote}}
}
\concept{Work Organizations}
\keyword{variable}
