% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_144.R
\name{fringe5}
\alias{fringe5}
\title{Maternity leave with full re-employment rights}
\source{
General Social Survey https://gss.norc.org
}
\description{
fringe5
}
\details{
Question I am going to read a list of fringe benefits that workers sometimes get in addition to their wages. Whether you receive it or not, please tell me whether you are eligible to receive each fringe benefit.
E. Maternity or paternity leave with full re-employment rights.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1357/vshow}{GSS Data Explorer page for \code{fringe5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 468 \tab 358 \tab 71 \tab 15 \tab 1517 \cr
   Total \tab 605 \tab 468 \tab 358 \tab 71 \tab 15 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{fringe1}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe6}},
\code{\link{othpay}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}}
}
\concept{Work Organizations}
\keyword{variable}
