% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{fringe9}
\alias{fringe9}
\title{Profit sharing or stock option program}
\source{
General Social Survey https://gss.norc.org
}
\description{
fringe9
}
\details{
Question I am going to read a list of fringe benefits that workers sometimes get in addition to their wages. Whether you receive it or not, please tell me whether you are eligible to receive each fringe benefit.
I. Profit sharing or stock option program.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1361/vshow}{GSS Data Explorer page for \code{fringe9}}.
}

\seealso{
Other Work Organizations: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{jobvshme}},
\code{\link{schoolng}},
\code{\link{unmanrel}},
\code{\link{unvote}}

Other Job: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{jobvshme}},
\code{\link{schoolng}},
\code{\link{unmanrel}}
}
\concept{Job}
\concept{Work Organizations}
\keyword{variable}
