% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{fringeok}
\alias{fringeok}
\title{Fringe benefits are good}
\source{
General Social Survey https://gss.norc.org
}
\description{
fringeok
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
F. My fringe benefits are good
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2801/vshow}{GSS Data Explorer page for \code{fringeok}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 705 \tab 548 \tab 214 \tab 304 \tab 10 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 636 \tab 539 \tab 211 \tab 301 \tab 31 \tab 16 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 416 \tab 389 \tab 138 \tab 202 \tab 13 \tab 29 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 466 \tab 441 \tab 162 \tab 163 \tab 9 \tab 9 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 536 \tab 520 \tab 156 \tab 176 \tab 26 \tab 34 \tab - \tab 2348 \cr
   2022 \tab 1553 \tab 639 \tab 752 \tab 266 \tab 253 \tab 37 \tab 18 \tab 26 \tab 3544 \cr
   Total \tab 8343 \tab 3398 \tab 3189 \tab 1147 \tab 1399 \tab 126 \tab 121 \tab 26 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{cowrkint}},
\code{\link{haveinfo}},
\code{\link{hlpequip}},
\code{\link{jobsecok}},
\code{\link{opdevel}},
\code{\link{promtefr}},
\code{\link{supcares}},
\code{\link{wkfreedm}}
}
\concept{Quality of Working Life}
\keyword{variable}
