% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{frinum}
\alias{frinum}
\title{Number of close friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
frinum
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to… How many close friends would you say you have?
}
\section{Values}{

\itemize{
\item \code{96} 96 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4162/vshow}{GSS Data Explorer page for \code{frinum}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 14 \tab 15 \tab 16 \tab 2 \tab 20 \tab 22 \tab 24 \tab 25 \tab 28 \tab 3 \tab 30 \tab 32 \tab 35 \tab 4 \tab 40 \tab 45 \tab 5 \tab 50 \tab 55 \tab 6 \tab 60 \tab 7 \tab 70 \tab 75 \tab 8 \tab 9 \tab 90 \tab 96 or more \tab don't know \tab Total \cr
   1986 \tab - \tab 77 \tab 95 \tab 145 \tab 55 \tab 2 \tab 25 \tab 1 \tab 174 \tab 55 \tab 1 \tab 1 \tab 14 \tab 1 \tab 180 \tab 17 \tab 1 \tab 1 \tab 186 \tab 3 \tab 1 \tab 171 \tab 14 \tab 1 \tab 149 \tab 1 \tab 30 \tab 1 \tab 1 \tab 49 \tab 4 \tab 1 \tab 10 \tab 3 \tab 1470 \cr
   Total \tab 0 \tab 77 \tab 95 \tab 145 \tab 55 \tab 2 \tab 25 \tab 1 \tab 174 \tab 55 \tab 1 \tab 1 \tab 14 \tab 1 \tab 180 \tab 17 \tab 1 \tab 1 \tab 186 \tab 3 \tab 1 \tab 171 \tab 14 \tab 1 \tab 149 \tab 1 \tab 30 \tab 1 \tab 1 \tab 49 \tab 4 \tab 1 \tab 10 \tab 3 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
