% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{frisex}
\alias{frisex}
\title{Sex of closest friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
frisex
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to…
C. Now thinking of your best friend, or the friend you feel closest to: Is this friend a man or a woman?
}
\section{Values}{

\itemize{
\item \code{1} man
\item \code{2} woman
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4165/vshow}{GSS Data Explorer page for \code{frisex}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
