% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{fritime}
\alias{fritime}
\title{Travel time to closest friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
fritime
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to…
E. About how long would it take you to get to where this friend lives? Think of the time it usually takes door to door.
}
\section{Values}{

\itemize{
\item \verb{[1]} lt 15 min
\item \verb{[2]} 15-30 min
\item \verb{[3]} 30 min-1 hr
\item \verb{[4]} 1-2 hr
\item \verb{[5]} 2-3 hr
\item \verb{[6]} 3-5 hr
\item \verb{[7]} 5-12 hr
\item \verb{[8]} 12+ hr
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4167/vshow}{GSS Data Explorer page for \code{fritime}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab lt 15 min \tab 15-30 min \tab 30 min-1 hr \tab 1-2 hr \tab 2-3 hr \tab 3-5 hr \tab 5-12 hr \tab 12+ hr \tab no answer \tab Total \cr
   1986 \tab 92 \tab 625 \tab 343 \tab 169 \tab 68 \tab 25 \tab 37 \tab 37 \tab 62 \tab 12 \tab 1470 \cr
   Total \tab 92 \tab 625 \tab 343 \tab 169 \tab 68 \tab 25 \tab 37 \tab 37 \tab 62 \tab 12 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
