% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{friwork}
\alias{friwork}
\title{Number of close friends are coworkers}
\source{
General Social Survey https://gss.norc.org
}
\description{
friwork
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to…
A. How many of these close friends are people you work with now?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4163/vshow}{GSS Data Explorer page for \code{friwork}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 14 \tab 15 \tab 16 \tab 2 \tab 20 \tab 22 \tab 25 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab Total \cr
   1986 \tab 77 \tab 950 \tab 178 \tab 9 \tab 2 \tab 2 \tab 4 \tab 2 \tab 98 \tab 2 \tab 1 \tab 2 \tab 58 \tab 1 \tab 19 \tab 26 \tab 11 \tab 4 \tab 9 \tab 1 \tab 14 \tab 1470 \cr
   Total \tab 77 \tab 950 \tab 178 \tab 9 \tab 2 \tab 2 \tab 4 \tab 2 \tab 98 \tab 2 \tab 1 \tab 2 \tab 58 \tab 1 \tab 19 \tab 26 \tab 11 \tab 4 \tab 9 \tab 1 \tab 14 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
