% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{frndawk}
\alias{frndawk}
\title{Discussing price is awkward among friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndawk
}
\details{
Question When friends buy or sell something to friends, discussing price is usually awkward or puts somebody on the spot.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1864/vshow}{GSS Data Explorer page for \code{frndawk}}.
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
