% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_119.R
\name{frndcon2}
\alias{frndcon2}
\title{Does friend 2 belong to r's congregation}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndcon2
}
\details{
Question Many people have some good friends they feel close to. Who are your good friends (other than your spouse)?  Just tell me their first names.
(LIST NAMES IN GRID BELOW). Is there anyone else?
NAME 1

Now let's go back and talk about (NAME). (REPEAT FOR EACH FRIEND)
Is (NAME) a member of your congregation?
2. Second Person Member of Respondent's Congregation
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} r isnt in congregtn
\item \verb{[4]} friend, no info
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1104/vshow}{GSS Data Explorer page for \code{frndcon2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab r isnt in congregtn \tab don't know \tab no answer \tab friend, no info \tab Total \cr
   1988 \tab 160 \tab 320 \tab 714 \tab 228 \tab 14 \tab 45 \tab - \tab 1481 \cr
   1998 \tab 1621 \tab 243 \tab 668 \tab 227 \tab 13 \tab 57 \tab 3 \tab 2832 \cr
   Total \tab 1781 \tab 563 \tab 1382 \tab 455 \tab 27 \tab 102 \tab 3 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{frndcon1}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}}
}
\concept{Religion}
\keyword{variable}
