% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_120.R
\name{frndden1}
\alias{frndden1}
\title{What is friend 1s specific denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndden1
}
\details{
Question What specific denomination is that, if any?
(PROBE FOR EXACT DENOMINATION. REFER TO PROTESTANT DENOMINATION LIST AND CIRCLE APPROPRIATE CODE BELOW).
\enumerate{
\item Denomination of First Person
}
}
\section{Values}{

\itemize{
\item \code{10} am baptist asso
\item \code{11} am bapt ch in usa
\item \code{12} nat bapt conv of am
\item \code{13} nat bapt conv usa
\item \code{14} southern baptist
\item \code{15} other baptists
\item \code{18} baptist-dk which
\item \code{20} afr meth episcopal
\item \code{21} afr meth ep zion
\item \code{22} united methodist
\item \code{23} other methodist
\item \code{28} methodist-dk which
\item \code{30} am lutheran
\item \code{31} luth ch in america
\item \code{32} lutheran-mo synod
\item \code{33} wi evan luth synod
\item \code{34} other lutheran
\item \code{35} evangelical lutheran
\item \code{38} lutheran-dk which
\item \code{40} presbyterian c in us
\item \code{41} united pres ch in usa
\item \code{42} other presbyterian
\item \code{43} presbyt ch usa
\item \code{48} presbyterian-dk wh
\item \code{50} episcopal
\item \code{60} other
\item \code{70} no denomination
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1113/vshow}{GSS Data Explorer page for \code{frndden1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab am baptist asso \tab am bapt ch in usa \tab nat bapt conv usa \tab southern baptist \tab other baptists \tab baptist-dk which \tab afr meth episcopal \tab afr meth ep zion \tab united methodist \tab other methodist \tab methodist-dk which \tab am lutheran \tab luth ch in america \tab lutheran-mo synod \tab other lutheran \tab evangelical lutheran \tab lutheran-dk which \tab presbyterian c in us \tab united pres ch in usa \tab presbyterian-dk wh \tab episcopal \tab other \tab no denomination \tab don't know \tab no answer \tab nat bapt conv of am \tab other presbyterian \tab presbyt ch usa \tab Total \cr
   1988 \tab 731 \tab 16 \tab 5 \tab 5 \tab 100 \tab 20 \tab 75 \tab 6 \tab 1 \tab 63 \tab 4 \tab 24 \tab 18 \tab 2 \tab 12 \tab 2 \tab 2 \tab 25 \tab 6 \tab 13 \tab 19 \tab 16 \tab 120 \tab 82 \tab 21 \tab 93 \tab - \tab - \tab - \tab 1481 \cr
   1998 \tab 2386 \tab 8 \tab 1 \tab 2 \tab 43 \tab 13 \tab 61 \tab 3 \tab 1 \tab 21 \tab 2 \tab 12 \tab 6 \tab 2 \tab 9 \tab 4 \tab 1 \tab 17 \tab 1 \tab 1 \tab 15 \tab 9 \tab 50 \tab 51 \tab 3 \tab 105 \tab 3 \tab 1 \tab 1 \tab 2832 \cr
   Total \tab 3117 \tab 24 \tab 6 \tab 7 \tab 143 \tab 33 \tab 136 \tab 9 \tab 2 \tab 84 \tab 6 \tab 36 \tab 24 \tab 4 \tab 21 \tab 6 \tab 3 \tab 42 \tab 7 \tab 14 \tab 34 \tab 25 \tab 170 \tab 133 \tab 24 \tab 198 \tab 3 \tab 1 \tab 1 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}}
}
\concept{Religion}
\keyword{variable}
