% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_120.R
\name{frndden2}
\alias{frndden2}
\title{What is friend 2s specific denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndden2
}
\details{
Question What specific denomination is that, if any?
(PROBE FOR EXACT DENOMINATION. REFER TO PROTESTANT DENOMINATION LIST AND CIRCLE APPROPRIATE CODE BELOW).
2. Denomination of Second Person
}
\section{Values}{

\itemize{
\item \verb{[10]} am baptist asso
\item \verb{[11]} am bapt ch in usa
\item \verb{[12]} nat bapt conv of am
\item \verb{[13]} nat bapt conv usa
\item \verb{[14]} southern baptist
\item \verb{[15]} other baptists
\item \verb{[18]} baptist-dk which
\item \verb{[20]} afr meth episcopal
\item \verb{[21]} afr meth ep zion
\item \verb{[22]} united methodist
\item \verb{[23]} other methodist
\item \verb{[28]} methodist-dk which
\item \verb{[30]} am lutheran
\item \verb{[31]} luth ch in america
\item \verb{[32]} lutheran-mo synod
\item \verb{[33]} wi evan luth synod
\item \verb{[34]} other lutheran
\item \verb{[35]} evangelical lutheran
\item \verb{[38]} lutheran-dk which
\item \verb{[40]} presbyterian c in us
\item \verb{[41]} united pres ch in usa
\item \verb{[42]} other presbyterian
\item \verb{[43]} presbyt ch usa
\item \verb{[48]} presbyterian-dk wh
\item \verb{[50]} episcopal
\item \verb{[60]} other
\item \verb{[70]} no denomination
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1114/vshow}{GSS Data Explorer page for \code{frndden2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab am baptist asso \tab am bapt ch in usa \tab nat bapt conv usa \tab southern baptist \tab other baptists \tab baptist-dk which \tab afr meth episcopal \tab afr meth ep zion \tab united methodist \tab other methodist \tab methodist-dk which \tab am lutheran \tab luth ch in america \tab lutheran-mo synod \tab other lutheran \tab evangelical lutheran \tab lutheran-dk which \tab presbyterian c in us \tab united pres ch in usa \tab other presbyterian \tab presbyterian-dk wh \tab episcopal \tab other \tab no denomination \tab don't know \tab no answer \tab nat bapt conv of am \tab presbyt ch usa \tab Total \cr
   1988 \tab 822 \tab 10 \tab 5 \tab 4 \tab 70 \tab 23 \tab 56 \tab 4 \tab 1 \tab 52 \tab 3 \tab 16 \tab 18 \tab 4 \tab 9 \tab 3 \tab 1 \tab 27 \tab 6 \tab 11 \tab 5 \tab 23 \tab 17 \tab 106 \tab 72 \tab 21 \tab 92 \tab - \tab - \tab 1481 \cr
   1998 \tab 2438 \tab 5 \tab 2 \tab - \tab 41 \tab 7 \tab 41 \tab 3 \tab - \tab 17 \tab 2 \tab 9 \tab 10 \tab - \tab 6 \tab 3 \tab 2 \tab 12 \tab 1 \tab 1 \tab 2 \tab 3 \tab 16 \tab 30 \tab 57 \tab 6 \tab 115 \tab 2 \tab 1 \tab 2832 \cr
   Total \tab 3260 \tab 15 \tab 7 \tab 4 \tab 111 \tab 30 \tab 97 \tab 7 \tab 1 \tab 69 \tab 5 \tab 25 \tab 28 \tab 4 \tab 15 \tab 6 \tab 3 \tab 39 \tab 7 \tab 12 \tab 7 \tab 26 \tab 33 \tab 136 \tab 129 \tab 27 \tab 207 \tab 2 \tab 1 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndden1}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}}
}
\concept{Religion}
\keyword{variable}
