% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_120.R
\name{frndden5}
\alias{frndden5}
\title{What is friend 5s specific denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndden5
}
\details{
Question What specific denomination is that, if any?
(PROBE FOR EXACT DENOMINATION. REFER TO PROTESTANT DENOMINATION LIST AND CIRCLE APPROPRIATE CODE BELOW).
5. Denomination of Fifth Person
}
\section{Values}{

\itemize{
\item \verb{[10]} am baptist asso
\item \verb{[11]} am bapt ch in usa
\item \verb{[12]} nat bapt conv of am
\item \verb{[13]} nat bapt conv usa
\item \verb{[14]} southern baptist
\item \verb{[15]} other baptists
\item \verb{[18]} baptist-dk which
\item \verb{[20]} afr meth episcopal
\item \verb{[21]} afr meth ep zion
\item \verb{[22]} united methodist
\item \verb{[23]} other methodist
\item \verb{[28]} methodist-dk which
\item \verb{[30]} am lutheran
\item \verb{[31]} luth ch in america
\item \verb{[32]} lutheran-mo synod
\item \verb{[33]} wi evan luth synod
\item \verb{[34]} other lutheran
\item \verb{[35]} evangelical lutheran
\item \verb{[38]} lutheran-dk which
\item \verb{[40]} presbyterian c in us
\item \verb{[41]} united pres ch in usa
\item \verb{[42]} other presbyterian
\item \verb{[43]} presbyt ch usa
\item \verb{[48]} presbyterian-dk wh
\item \verb{[50]} episcopal
\item \verb{[60]} other
\item \verb{[70]} no denomination
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1117/vshow}{GSS Data Explorer page for \code{frndden5}}.

Counts by year:\tabular{lllllllllllllllllllllllllll}{
   year \tab iap \tab am baptist asso \tab am bapt ch in usa \tab southern baptist \tab other baptists \tab baptist-dk which \tab afr meth episcopal \tab united methodist \tab other methodist \tab methodist-dk which \tab am lutheran \tab luth ch in america \tab lutheran-mo synod \tab wi evan luth synod \tab other lutheran \tab lutheran-dk which \tab presbyterian c in us \tab united pres ch in usa \tab presbyt ch usa \tab presbyterian-dk wh \tab episcopal \tab other \tab no denomination \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2625 \tab 5 \tab 1 \tab 12 \tab 4 \tab 16 \tab 1 \tab 3 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 9 \tab 14 \tab 34 \tab 3 \tab 83 \tab 2832 \cr
   Total \tab 2625 \tab 5 \tab 1 \tab 12 \tab 4 \tab 16 \tab 1 \tab 3 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 9 \tab 14 \tab 34 \tab 3 \tab 83 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}}
}
\concept{Religion}
\keyword{variable}
