% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{frndfam}
\alias{frndfam}
\title{Friends or family more important}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndfam
}
\details{
Question While some people think that close relatives should be the most important people in one"s life, others think that it should be close friends who are the most important. Which view is closer to your opinion? Please choose a number on a scale of 1 to 7 whe
}
\section{Values}{

\itemize{
\item \code{1} close relatives should definitely be the most important
\item \code{4} both should be equally important
\item \code{7} close friends should definitely be the most important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7782/vshow}{GSS Data Explorer page for \code{frndfam}}.
}

\keyword{variable}
