% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_121.R
\name{frndfnd4}
\alias{frndfnd4}
\title{Fund.-lib. of 4th friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndfnd4
}
\details{
Question Fundamentalism/Liberalism of Friend''s Religion
4. Fourth Person''s Fundamentalism/Liberalism
}
\section{Values}{

\itemize{
\item \verb{[1]} fund
\item \verb{[2]} moderate
\item \verb{[3]} liberal
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1126/vshow}{GSS Data Explorer page for \code{frndfnd4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab fund \tab moderate \tab liberal \tab no answer \tab Total \cr
   1998 \tab 2046 \tab 78 \tab 190 \tab 98 \tab 420 \tab 2832 \cr
   Total \tab 2046 \tab 78 \tab 190 \tab 98 \tab 420 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd5}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}}
}
\concept{Religion}
\keyword{variable}
