% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{frndknow}
\alias{frndknow}
\title{How many of r's friends know each other?}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndknow
}
\details{
Question Some people have friends who mostly know one another. Other people have friends who don't know one another. would you say that all of your friends know one another, most of your friends know one another, only a few of your friends know one another, or none of your friends know one another?
}
\section{Values}{

\itemize{
\item \code{1} all know each other
\item \code{2} most know each other
\item \code{3} few know each other
\item \code{4} none know each other
\item \code{5} no friends(vol.)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/962/vshow}{GSS Data Explorer page for \code{frndknow}}.
}

\seealso{
Other Social Networks: 
\code{\link{talkpol}},
\code{\link{talkpol3}}
}
\concept{Social Networks}
\keyword{variable}
