% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_121.R
\name{frndoth3}
\alias{frndoth3}
\title{What is friend 3s other prot denom}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndoth3
}
\details{
Question What specific denomination is that, if any?
3. Denomination of third friend
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1120/vshow}{GSS Data Explorer page for \code{frndoth3}}.
}

\seealso{
Other Religion: 
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}}
}
\concept{Religion}
\keyword{variable}
