% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_121.R
\name{frndoth4}
\alias{frndoth4}
\title{What is friend 4s other prot denom}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndoth4
}
\details{
Question What specific denomination is that, if any?
4. Denomination of fourth friend
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1121/vshow}{GSS Data Explorer page for \code{frndoth4}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 138 \tab 150 \tab 3 \tab 31 \tab 33 \tab 35 \tab 36 \tab 40 \tab 56 \tab 58 \tab 64 \tab 68 \tab 69 \tab 70 \tab 77 \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2710 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab 96 \tab 2832 \cr
   Total \tab 2710 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab 96 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth3}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}}
}
\concept{Religion}
\keyword{variable}
