% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_121.R
\name{frndrac1}
\alias{frndrac1}
\title{What is friend 1s race}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndrac1
}
\details{
Question Is (NAME) Asian, Black, Hispanic, White or something else?
\enumerate{
\item Race of First Friend
}
}
\section{Values}{

\itemize{
\item \verb{[1]} asian
\item \verb{[2]} black
\item \verb{[3]} hispanic
\item \verb{[4]} white
\item \verb{[5]} something else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1128/vshow}{GSS Data Explorer page for \code{frndrac1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab asian \tab black \tab hispanic \tab white \tab something else \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1454 \tab 37 \tab 172 \tab 89 \tab 979 \tab 23 \tab 1 \tab 77 \tab 2832 \cr
   Total \tab 1454 \tab 37 \tab 172 \tab 89 \tab 979 \tab 23 \tab 1 \tab 77 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac2}}
}
\concept{Religion}
\keyword{variable}
