% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{frndrac3}
\alias{frndrac3}
\title{What is friend 3s race}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndrac3
}
\details{
Question Is (NAME) Asian, Black, Hispanic, White or something else?
3. Race of Third Friend
}
\section{Values}{

\itemize{
\item \code{1} asian
\item \code{2} black
\item \code{3} hispanic
\item \code{4} white
\item \code{5} something else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1130/vshow}{GSS Data Explorer page for \code{frndrac3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab asian \tab black \tab hispanic \tab white \tab something else \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1853 \tab 21 \tab 104 \tab 46 \tab 723 \tab 16 \tab 1 \tab 68 \tab 2832 \cr
   Total \tab 1853 \tab 21 \tab 104 \tab 46 \tab 723 \tab 16 \tab 1 \tab 68 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numbwfrd}},
\code{\link{numfrend}}
}
\concept{Religion}
\keyword{variable}
