% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{frndrac4}
\alias{frndrac4}
\title{What is friend 4s race}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndrac4
}
\details{
Question Is (NAME) Asian, Black, Hispanic, White or something else?
4. Race of Fourth Friend
}
\section{Values}{

\itemize{
\item \verb{[1]} asian
\item \verb{[2]} black
\item \verb{[3]} hispanic
\item \verb{[4]} white
\item \verb{[5]} something else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1131/vshow}{GSS Data Explorer page for \code{frndrac4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab asian \tab black \tab hispanic \tab white \tab something else \tab no answer \tab Total \cr
   1998 \tab 2046 \tab 23 \tab 82 \tab 36 \tab 569 \tab 10 \tab 66 \tab 2832 \cr
   Total \tab 2046 \tab 23 \tab 82 \tab 36 \tab 569 \tab 10 \tab 66 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numbwfrd}},
\code{\link{numfrend}}
}
\concept{Religion}
\keyword{variable}
