% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{frndrac5}
\alias{frndrac5}
\title{What is friend 5s race}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndrac5
}
\details{
Question Is (NAME) Asian, Black, Hispanic, White or something else?
5. Race of Fifth Friend
}
\section{Values}{

\itemize{
\item \verb{[1]} asian
\item \verb{[2]} black
\item \verb{[3]} hispanic
\item \verb{[4]} white
\item \verb{[5]} something else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1132/vshow}{GSS Data Explorer page for \code{frndrac5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab asian \tab black \tab hispanic \tab white \tab something else \tab no answer \tab Total \cr
   1998 \tab 2212 \tab 14 \tab 49 \tab 34 \tab 453 \tab 6 \tab 64 \tab 2832 \cr
   Total \tab 2212 \tab 14 \tab 49 \tab 34 \tab 453 \tab 6 \tab 64 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numbwfrd}},
\code{\link{numfrend}}
}
\concept{Religion}
\keyword{variable}
