% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_120.R
\name{frndrel3}
\alias{frndrel3}
\title{What is friend 3s religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndrel3
}
\details{
Question What''s is (NAME''S) religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?
3. Religion of Third Person
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1110/vshow}{GSS Data Explorer page for \code{frndrel3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1988 \tab 336 \tab 497 \tab 248 \tab 36 \tab 85 \tab 36 \tab 167 \tab 76 \tab 1481 \cr
   1998 \tab 2047 \tab 274 \tab 150 \tab 19 \tab 71 \tab 33 \tab 161 \tab 77 \tab 2832 \cr
   Total \tab 2383 \tab 771 \tab 398 \tab 55 \tab 156 \tab 69 \tab 328 \tab 153 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndrel4}},
\code{\link{frndrel5}}
}
\concept{Religion}
\keyword{variable}
