% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_120.R
\name{frndrel4}
\alias{frndrel4}
\title{What is friend 4s religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndrel4
}
\details{
Question What''s is (NAME''S) religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?
4. Religion of Fourth Person
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1111/vshow}{GSS Data Explorer page for \code{frndrel4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2197 \tab 201 \tab 127 \tab 16 \tab 52 \tab 31 \tab 132 \tab 76 \tab 2832 \cr
   Total \tab 2197 \tab 201 \tab 127 \tab 16 \tab 52 \tab 31 \tab 132 \tab 76 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndrel3}},
\code{\link{frndrel5}}
}
\concept{Religion}
\keyword{variable}
