% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_605.R
\name{frstgrp3}
\alias{frstgrp3}
\title{Which group meet name3}
\source{
General Social Survey https://gss.norc.org
}
\description{
frstgrp3
}
\details{
Question What are the groups you both belong to? Circle all that apply.
}
\section{Values}{

\itemize{
\item \code{1} fraternal groups
\item \code{2} service clubs
\item \code{3} veterans groups
\item \code{4} political clubs
\item \code{5} labor unions
\item \code{6} sports groups
\item \code{7} youth groups
\item \code{8} school service groups
\item \code{9} hobby or garden clubs
\item \code{10} school fraternities or sororities
\item \code{11} nationality groups
\item \code{12} farm organizations
\item \code{13} literary art discussion or study grps
\item \code{14} professional or academic societies
\item \code{15} church-affiliated groups
\item \code{16} other groups
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5547/vshow}{GSS Data Explorer page for \code{frstgrp3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab nationality groups \tab literary art discussion or study grps \tab professional or academic societies \tab church-affiliated groups \tab other groups \tab no answer \tab Total \cr
   2004 \tab 2741 \tab 1 \tab 4 \tab 2 \tab 12 \tab 4 \tab 48 \tab 2812 \cr
   Total \tab 2741 \tab 1 \tab 4 \tab 2 \tab 12 \tab 4 \tab 48 \tab 2812 \cr
}
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp5d}},
\code{\link{bthgrp5e}},
\code{\link{bthgrp5f}},
\code{\link{frstgrp1}},
\code{\link{frstgrp2}},
\code{\link{frstgrp4}},
\code{\link{frstgrp5}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
