% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_365.R
\name{frstmet1}
\alias{frstmet1}
\title{R first met the named alter in the group}
\source{
General Social Survey https://gss.norc.org
}
\description{
frstmet1
}
\details{
Question Did you first meet (NAME) in one of the groups in the previous question, or did you first meet somewhere else?
A. Did you meet the first name in one of these groups or did you meet somewhere else?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, one of these groups
\item \verb{[2]} no, somewhere else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3303/vshow}{GSS Data Explorer page for \code{frstmet1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes, one of these groups \tab no, somewhere else \tab no answer \tab Total \cr
   2004 \tab 2674 \tab 24 \tab 62 \tab 52 \tab 2812 \cr
   Total \tab 2674 \tab 24 \tab 62 \tab 52 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
