% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_366.R
\name{frstmet3}
\alias{frstmet3}
\title{R first met the named alter in the group}
\source{
General Social Survey https://gss.norc.org
}
\description{
frstmet3
}
\details{
Question Did you first meet (NAME) in one of the groups in the previous question, or did you first meet somewhere else?
C. Did you meet the third name in one of these groups or did you meet somewhere else?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, one of these groups
\item \verb{[2]} no, somewhere else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3305/vshow}{GSS Data Explorer page for \code{frstmet3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, one of these groups \tab no, somewhere else \tab no answer \tab not available in this year \tab Total \cr
   2004 \tab 2718 \tab 23 \tab 23 \tab 48 \tab - \tab 2812 \cr
   Total \tab 2718 \tab 23 \tab 23 \tab 48 \tab 0 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{churchtx}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{infrmgrp}},
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{mtfrat4}},
\code{\link{mtgreek1}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mthobby1}},
\code{\link{mthobby2}},
\code{\link{mthobby3}},
\code{\link{mtinfrm1}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtlit3}},
\code{\link{mtlit4}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}},
\code{\link{mtother1}},
\code{\link{mtother2}},
\code{\link{mtother3}},
\code{\link{mtother4}},
\code{\link{mtother5}},
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}},
\code{\link{mtschl1}},
\code{\link{mtschl2}},
\code{\link{mtschl3}},
\code{\link{mtschl4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv7}},
\code{\link{mtserv8}},
\code{\link{mtsport1}},
\code{\link{mtsport2}},
\code{\link{mtsport3}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtunion4}},
\code{\link{mtvet1}},
\code{\link{mtvet2}},
\code{\link{mtvet3}},
\code{\link{mtvet4}},
\code{\link{mtvet5}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}},
\code{\link{mtyouth4}},
\code{\link{mtyouth5}},
\code{\link{numchurh}},
\code{\link{numfarm}},
\code{\link{numfrat}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numinfrm}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numpolit}},
\code{\link{numprof}},
\code{\link{numschl}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{relneg}},
\code{\link{typchurh}},
\code{\link{typfarm}},
\code{\link{typfrat}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typinfrm}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typother}},
\code{\link{typpolit}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typserv}},
\code{\link{typsport}},
\code{\link{typunion}},
\code{\link{typvet}},
\code{\link{typyouth}},
\code{\link{yrchurh1}},
\code{\link{yrchurh2}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrfarm3}},
\code{\link{yrfrat1}},
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby1}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrnat1}},
\code{\link{yrnat2}},
\code{\link{yrnat3}},
\code{\link{yrother1}},
\code{\link{yrother2}},
\code{\link{yrother3}},
\code{\link{yrother4}},
\code{\link{yrother5}},
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}},
\code{\link{yrprof6}},
\code{\link{yrschl1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yrserv1}},
\code{\link{yrserv2}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}},
\code{\link{yrserv6}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}},
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrsport6}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}

Other Social Networks: 
\code{\link{acqasian}},
\code{\link{acqattnd}},
\code{\link{acqblack}},
\code{\link{acqbrnda}},
\code{\link{acqchild}},
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqdems}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhme}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmno}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqfmwht}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjans}},
\code{\link{acqjose}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqlinda}},
\code{\link{acqmaria}},
\code{\link{acqmark}},
\code{\link{acqmils}},
\code{\link{acqmyrac}},
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhlib}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}},
\code{\link{acqnhwht}},
\code{\link{acqnoatt}},
\code{\link{acqprisn}},
\code{\link{acqrachl}},
\code{\link{acqreps}},
\code{\link{acqshawn}},
\code{\link{acqsocs}},
\code{\link{acqunemp}},
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvalib}},
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhme}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkno}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}},
\code{\link{acqwkwht}},
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{comamind}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{fricall}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndknow}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{grpother}},
\code{\link{grpsprts}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{hlpinflu}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{lentto}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{peocntct}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkedto}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}},
\code{\link{trtasian}},
\code{\link{trtattnd}},
\code{\link{trtblack}},
\code{\link{trtbrnda}},
\code{\link{trtchild}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtcops}},
\code{\link{trtdems}},
\code{\link{trtelecs}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtjose}},
\code{\link{trtkaren}},
\code{\link{trtkeith}},
\code{\link{trtkevin}},
\code{\link{trtlaws}},
\code{\link{trtlib}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtmils}},
\code{\link{trtmyrac}},
\code{\link{trtnoatt}},
\code{\link{trtprisn}},
\code{\link{trtrachl}},
\code{\link{trtreps}},
\code{\link{trtshawn}},
\code{\link{trtsocs}},
\code{\link{trtunemp}},
\code{\link{trtwhite}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{youinflu}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Social Networks}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
