% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{frterror}
\alias{frterror}
\title{Terrorism by foreigner's more threat 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
frterror
}
\details{
Question Are the following threats to the United States greater, about the same, or less today than they were 10 years ago?
D. Terrorism by foreigners
}
\section{Values}{

\itemize{
\item \verb{[1]} greater
\item \verb{[2]} about the same
\item \verb{[3]} less
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5260/vshow}{GSS Data Explorer page for \code{frterror}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab greater \tab about the same \tab less \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2000 \tab 1398 \tab 899 \tab 336 \tab 79 \tab 83 \tab 22 \tab - \tab 2817 \cr
   Total \tab 1398 \tab 899 \tab 336 \tab 79 \tab 83 \tab 22 \tab 0 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askcrime}},
\code{\link{askdrink}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{askmentl}},
\code{\link{asksexor}},
\code{\link{bugging}},
\code{\link{camwrk}},
\code{\link{chkfinan}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{chktravl}},
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{compfin}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{lietest}},
\code{\link{nuclrwar}},
\code{\link{punarms}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}},
\code{\link{rptcowrk}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secdocs}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{sectech}},
\code{\link{secterr}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{srchwrk}},
\code{\link{takearms}},
\code{\link{taketrck}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{testdrug}},
\code{\link{usspy}},
\code{\link{usterror}}

Other Terrorism: 
\code{\link{done911a}},
\code{\link{done911b}},
\code{\link{done911c}},
\code{\link{done911d}},
\code{\link{done911e}},
\code{\link{done911f}},
\code{\link{done911g}},
\code{\link{done911h}},
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911l}},
\code{\link{done911m}},
\code{\link{done911n}},
\code{\link{doneelse}},
\code{\link{eff911a}},
\code{\link{eff911b}},
\code{\link{eff911c}},
\code{\link{eff911d}},
\code{\link{eff911e}},
\code{\link{eff911f}},
\code{\link{eff911g}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}},
\code{\link{eff911m}},
\code{\link{eff911n}},
\code{\link{secterr}},
\code{\link{usterror}}

Other United States: 
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amimp}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amprogrn}},
\code{\link{amproud}},
\code{\link{amproud1}},
\code{\link{amrank}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{askcomp}},
\code{\link{askcrime}},
\code{\link{askdrink}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{askmentl}},
\code{\link{asksexor}},
\code{\link{asnchng}},
\code{\link{belikeus}},
\code{\link{blkchng}},
\code{\link{bugging}},
\code{\link{camwrk}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{citizen}},
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{compfin}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{corrupt}},
\code{\link{decsorgs}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{demworks}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{engoffcl}},
\code{\link{excldimm}},
\code{\link{finanqs}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{freetrde}},
\code{\link{futenrgy}},
\code{\link{guerilla}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}},
\code{\link{hispchng}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immcrime}},
\code{\link{immcult}},
\code{\link{immfare}},
\code{\link{immideas}},
\code{\link{immjobs}},
\code{\link{imports}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intlincs}},
\code{\link{jewschng}},
\code{\link{landwar}},
\code{\link{leakinfo}},
\code{\link{lessnuke}},
\code{\link{lessprd}},
\code{\link{letin}},
\code{\link{letin1}},
\code{\link{lietest}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{morenuke}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{nonuke}},
\code{\link{notam}},
\code{\link{nuclrwar}},
\code{\link{nukewar}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{punarms}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}},
\code{\link{refugees}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{rptcowrk}},
\code{\link{satdemoc}},
\code{\link{secdocs}},
\code{\link{secprvcy}},
\code{\link{sectech}},
\code{\link{shortcom}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{srchwrk}},
\code{\link{stoprndm}},
\code{\link{takearms}},
\code{\link{taketrck}},
\code{\link{taphme}},
\code{\link{tapphone}},
\code{\link{tapwrk}},
\code{\link{testdrug}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{usamind}},
\code{\link{usangry}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{usdoenuf}},
\code{\link{usenthus}},
\code{\link{usfrustr}},
\code{\link{ushisp}},
\code{\link{ushopefl}},
\code{\link{usintl}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{ussat}},
\code{\link{usspy}},
\code{\link{usterror}},
\code{\link{usun}},
\code{\link{usupset}},
\code{\link{uswar}},
\code{\link{uswary}},
\code{\link{uswht}},
\code{\link{usworry}},
\code{\link{whtchng}},
\code{\link{wotrial}},
\code{\link{wrldgovt}}
}
\concept{National Security}
\concept{Terrorism}
\concept{United States}
\keyword{variable}
