% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_156.R
\name{fstspdeg}
\alias{fstspdeg}
\title{Highest degree of first spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
fstspdeg
}
\details{
Question RESPONDENT'S FIRST SPOUSE'S DEGREE
}
\section{Values}{

\itemize{
\item \code{0} less than high school
\item \code{1} high school
\item \code{2} associate/junior college
\item \code{3} bachelor's
\item \code{4} graduate
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1478/vshow}{GSS Data Explorer page for \code{fstspdeg}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab less than high school \tab high school \tab associate/junior college \tab bachelor's \tab graduate \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2432 \tab 119 \tab 309 \tab 20 \tab 48 \tab 24 \tab 34 \tab 6 \tab 2992 \cr
   Total \tab 2432 \tab 119 \tab 309 \tab 20 \tab 48 \tab 24 \tab 34 \tab 6 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{fstspedc}},
\code{\link{fstspyr}},
\code{\link{marnum}}
}
\concept{Family Mobility}
\keyword{variable}
