% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{fstspmed}
\alias{fstspmed}
\title{1st spouse's mothers highest yr of school}
\source{
General Social Survey https://gss.norc.org
}
\description{
fstspmed
}
\details{
Question What is the highest grade or year of regular school that your (first/former) (husband's/wife's) mother ever completed?
CODE EXACT GRADE.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1480/vshow}{GSS Data Explorer page for \code{fstspmed}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2387 \tab 5 \tab 5 \tab 3 \tab 138 \tab 3 \tab 13 \tab 2 \tab 14 \tab 3 \tab 2 \tab 1 \tab 1 \tab 6 \tab 5 \tab 2 \tab 1 \tab 2 \tab 29 \tab 2 \tab 311 \tab 57 \tab 2992 \cr
   Total \tab 2387 \tab 5 \tab 5 \tab 3 \tab 138 \tab 3 \tab 13 \tab 2 \tab 14 \tab 3 \tab 2 \tab 1 \tab 1 \tab 6 \tab 5 \tab 2 \tab 1 \tab 2 \tab 29 \tab 2 \tab 311 \tab 57 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{fstspped}},
\code{\link{spfam16}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}
}
\concept{Family Mobility}
\keyword{variable}
