% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_673.R
\name{ftruse}
\alias{ftruse}
\title{Likelihood r uses internet in next 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
ftruse
}
\details{
Question How likely is it that you will personally use the Internet in the next 12 months?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} rather likely
\item \verb{[3]} rather unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8144/vshow}{GSS Data Explorer page for \code{ftruse}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab very likely \tab rather likely \tab rather unlikely \tab very unlikely \tab can't choose \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 6 \tab 3 \tab 5 \tab 40 \tab 9 \tab 3245 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 6 \tab 3 \tab 5 \tab 40 \tab 9 \tab 3245 \tab 1 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intaccprxy}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint6}},
\code{\link{nouseint7}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{prxyuse}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
