% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{ftwork}
\alias{ftwork}
\title{Young should be employed full-time}
\source{
General Social Survey https://gss.norc.org
}
\description{
ftwork
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important is it for them to be...
To be employed full-time
}
\section{Values}{

\itemize{
\item \code{1} extremely important
\item \code{2} quite important
\item \code{3} somewhat important
\item \code{4} not too important
\item \code{5} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2902/vshow}{GSS Data Explorer page for \code{ftwork}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely important \tab quite important \tab somewhat important \tab not too important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 831 \tab 325 \tab 156 \tab 51 \tab 12 \tab 13 \tab 5 \tab 2765 \cr
   Total \tab 1372 \tab 831 \tab 325 \tab 156 \tab 51 \tab 12 \tab 13 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{ftwork1}},
\code{\link{ownhh}},
\code{\link{ownhh1}}
}
\concept{Adult Transitions}
\keyword{variable}
