% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{ftwork1}
\alias{ftwork1}
\title{Aged should be employed full-time}
\source{
General Social Survey https://gss.norc.org
}
\description{
ftwork1
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important is it for them to be...
To be employed full-time
A. By what age should this normally occur?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2903/vshow}{GSS Data Explorer page for \code{ftwork1}}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 35 \tab 50 \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1453 \tab 1 \tab 1 \tab 1 \tab 36 \tab 19 \tab 292 \tab 57 \tab 96 \tab 162 \tab 203 \tab 125 \tab 88 \tab 166 \tab 12 \tab 10 \tab 5 \tab 1 \tab 8 \tab 1 \tab 1 \tab 26 \tab 1 \tab 2765 \cr
   Total \tab 1453 \tab 1 \tab 1 \tab 1 \tab 36 \tab 19 \tab 292 \tab 57 \tab 96 \tab 162 \tab 203 \tab 125 \tab 88 \tab 166 \tab 12 \tab 10 \tab 5 \tab 1 \tab 8 \tab 1 \tab 1 \tab 26 \tab 1 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{ftwork}},
\code{\link{ownhh}},
\code{\link{ownhh1}}
}
\concept{Adult Transitions}
\keyword{variable}
