% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{fucitzn}
\alias{fucitzn}
\title{Is r planning/appling for us citizenship or not}
\source{
General Social Survey https://gss.norc.org
}
\description{
fucitzn
}
\details{
Question Are you…
}
\section{Values}{

\itemize{
\item \verb{[1]} currently applying for u.s. citizenship
\item \verb{[2]} planning to apply for u.s. citizenship
\item \verb{[3]} not planning to apply to u.s. citizenship
\item \verb{[4]} not eligible to become a u.s. citizen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5284/vshow}{GSS Data Explorer page for \code{fucitzn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab currently applying for u.s. citizenship \tab planning to apply for u.s. citizenship \tab not planning to apply to u.s. citizenship \tab not eligible to become a u.s. citizen \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 1884 \tab 24 \tab 61 \tab 38 \tab 9 \tab 1 \tab 6 \tab - \tab 2023 \cr
   2010 \tab 1920 \tab 6 \tab 76 \tab 34 \tab 4 \tab 2 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 1846 \tab 10 \tab 57 \tab 43 \tab 11 \tab 2 \tab 5 \tab - \tab 1974 \cr
   2014 \tab 2378 \tab 11 \tab 84 \tab 32 \tab 30 \tab - \tab 3 \tab - \tab 2538 \cr
   2016 \tab 2709 \tab 19 \tab 65 \tab 45 \tab 14 \tab 1 \tab 14 \tab - \tab 2867 \cr
   2018 \tab 2213 \tab 16 \tab 49 \tab 46 \tab 17 \tab 3 \tab 4 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3442 \tab 11 \tab 44 \tab 24 \tab 9 \tab 3 \tab 11 \tab - \tab 3544 \cr
   Total \tab 16392 \tab 97 \tab 436 \tab 262 \tab 94 \tab 12 \tab 45 \tab 4032 \tab 21370 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethadapt}},
\code{\link{ethclose}},
\code{\link{ethdiff}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethnic}},
\code{\link{ethnofit}},
\code{\link{ethnum}},
\code{\link{ethorgs}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hspgovt}},
\code{\link{meltpot}},
\code{\link{natborn}},
\code{\link{owneth}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{schleth}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{teacheth}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{whoteach}},
\code{\link{whtgovt}}

Other Citizenship: 
\code{\link{actassoc}},
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obeylaws}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{typevisa}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{visa}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Citizenship}
\concept{Ethnicity}
\concept{Immigration}
\keyword{variable}
