% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_664.R
\name{fund_next}
\alias{fund_next}
\title{Fundamentalist - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
fund_next
}
\details{
Question Fundamentalism/Liberalism of Respondent's Religion
}
\section{Values}{

\itemize{
\item \verb{[1]} fundamentalist
\item \verb{[2]} moderate
\item \verb{[3]} liberal
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8040/vshow}{GSS Data Explorer page for \code{fund_next}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab fundamentalist \tab moderate \tab liberal \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 199 \tab 430 \tab 524 \tab 2313 \tab 75 \tab 3 \tab 3544 \cr
   Total \tab 0 \tab 199 \tab 430 \tab 524 \tab 2313 \tab 75 \tab 3 \tab 3544 \cr
}
}

\concept{GSS Next}
\keyword{variable}
