% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{futpromo}
\alias{futpromo}
\title{How likely you will be promoted next five years}
\source{
General Social Survey https://gss.norc.org
}
\description{
futpromo
}
\details{
Question In the next five years, how likely are you to be promoted?  Would you say very likely, likely, not very likely, or not likely at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} likely
\item \verb{[3]} not very likely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1303/vshow}{GSS Data Explorer page for \code{futpromo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab likely \tab not very likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 201 \tab 170 \tab 181 \tab 296 \tab 38 \tab 26 \tab 1517 \cr
   Total \tab 605 \tab 201 \tab 170 \tab 181 \tab 296 \tab 38 \tab 26 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}
}
\concept{Work Organizations}
\keyword{variable}
