% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_267.R
\name{game12}
\alias{game12}
\title{R use web for games in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
game12
}
\details{
Question In the past 12 months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/02 have you used the Web to do each of these things:
J. Play a game with someone on another computer?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2424/vshow}{GSS Data Explorer page for \code{game12}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 138 \tab 526 \tab 78 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 187 \tab 802 \tab 94 \tab 1 \tab 2765 \cr
   Total \tab 3756 \tab 325 \tab 1328 \tab 172 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{chat12}},
\code{\link{econ12}},
\code{\link{health12}},
\code{\link{locate12}},
\code{\link{newjob12}},
\code{\link{people12}},
\code{\link{polinf12}},
\code{\link{travel12}}
}
\concept{Information Society}
\keyword{variable}
