% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{gasregs}
\alias{gasregs}
\title{How much r favor requiring to make car's using less gasoline}
\source{
General Social Survey https://gss.norc.org
}
\description{
gasregs
}
\details{
Question How much do you favor or oppose requiring car makers to make cars and trucks that use less gasoline? Do you strongly favor, favor, neither favor nor oppose, oppose, or strongly oppose such a requirement?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3365/vshow}{GSS Data Explorer page for \code{gasregs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 645 \tab 197 \tab 52 \tab 14 \tab 9 \tab 10 \tab 1 \tab 4510 \cr
   Total \tab 3582 \tab 645 \tab 197 \tab 52 \tab 14 \tab 9 \tab 10 \tab 1 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgw}},
\code{\link{bizinfgw}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
