% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{gavepol}
\alias{gavepol}
\title{Has r given money to candidate or cause}
\source{
General Social Survey https://gss.norc.org
}
\description{
gavepol
}
\details{
Question In the past three or four years, have you contributed money to a political party or candidate or to any other political cause?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1040/vshow}{GSS Data Explorer page for \code{gavepol}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1987 \tab - \tab 390 \tab 1417 \tab 12 \tab - \tab 1819 \cr
   Total \tab 0 \tab 390 \tab 1417 \tab 12 \tab 0 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{blkgains}},
\code{\link{blkinflu}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{fedtrust}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{givegrp}},
\code{\link{governor}},
\code{\link{intpol}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{mempolit}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{mostact}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{swayvote}},
\code{\link{usrep}},
\code{\link{workpol}},
\code{\link{writegrp}}

Other Political: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{anomia7}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{bribe}},
\code{\link{buspow}},
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf11}},
\code{\link{caninf12}},
\code{\link{caninf13}},
\code{\link{caninf14}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf6}},
\code{\link{caninf7}},
\code{\link{caninf8}},
\code{\link{caninf9}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{civic}},
\code{\link{cntctgov}},
\code{\link{congeth}},
\code{\link{contemp}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{corruptn}},
\code{\link{demworks}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{engballt}},
\code{\link{ethorgs}},
\code{\link{fixmistk}},
\code{\link{giveoth}},
\code{\link{govdook}},
\code{\link{governor}},
\code{\link{govngos}},
\code{\link{govtpow}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{interpol}},
\code{\link{intpol}},
\code{\link{joindem}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{leftrght}},
\code{\link{libtemp}},
\code{\link{localgvt}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{ntcitvte}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{owneth}},
\code{\link{partpol}},
\code{\link{partyid}},
\code{\link{polactve}},
\code{\link{polcamp}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff2}},
\code{\link{poleff20}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy3}},
\code{\link{polfrel}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polinfo}},
\code{\link{polint}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polnews}},
\code{\link{polpapr}},
\code{\link{polrally}},
\code{\link{polsfair}},
\code{\link{poltv}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{polwww}},
\code{\link{powerun}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prespop}},
\code{\link{refrndms}},
\code{\link{relcand}},
\code{\link{schlhead}},
\code{\link{servepeo}},
\code{\link{signdpet}},
\code{\link{swayvote}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usrep}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{workpol}}
}
\concept{Political}
\concept{Socio-political participation}
\keyword{variable}
