% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_582.R
\name{gender1}
\alias{gender1}
\title{Gender of 1st person}
\source{
General Social Survey https://gss.norc.org
}
\description{
gender1
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
2. Gender of first person
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5333/vshow}{GSS Data Explorer page for \code{gender1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab don't know \tab not available in this release \tab 3 \tab skipped on web \tab Total \cr
   1975 \tab - \tab 1210 \tab 280 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 755 \tab 572 \tab 172 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1220 \tab 310 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1168 \tab 364 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1126 \tab 342 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1324 \tab 536 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1193 \tab 401 \tab 5 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1069 \tab 396 \tab 8 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1130 \tab 404 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1055 \tab 415 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1241 \tab 576 \tab 2 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 1056 \tab 425 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 1065 \tab 458 \tab 14 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 916 \tab 446 \tab 10 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1046 \tab 471 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1094 \tab 512 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2063 \tab 927 \tab 2 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 1902 \tab 978 \tab 24 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 1832 \tab 989 \tab 11 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 1569 \tab 1243 \tab 5 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 1500 \tab 1265 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 1632 \tab 1179 \tab 1 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 2565 \tab 1945 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1157 \tab 865 \tab 1 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1114 \tab 925 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1131 \tab 843 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1431 \tab 1107 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 1510 \tab 1357 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 1227 \tab 1121 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 1779 \tab 908 \tab 849 \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 3544 \cr
   Total \tab 2534 \tab 39026 \tab 22101 \tab 88 \tab 1 \tab 4032 \tab 6 \tab 1 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away1}},
\code{\link{away2}},
\code{\link{gender2}},
\code{\link{mar1}},
\code{\link{mar2}},
\code{\link{old1}},
\code{\link{old2}},
\code{\link{relate2}},
\code{\link{where1}}
}
\concept{Household Members and Structure}
\keyword{variable}
