% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_582.R
\name{gender2}
\alias{gender2}
\title{Gender of 2nd person}
\source{
General Social Survey https://gss.norc.org
}
\description{
gender2
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
2. Gender of second person
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[3]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5339/vshow}{GSS Data Explorer page for \code{gender2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab not available in this release \tab skipped on web \tab Total \cr
   1975 \tab 203 \tab 100 \tab 1187 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 893 \tab 53 \tab 553 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 256 \tab 128 \tab 1146 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 295 \tab 121 \tab 1116 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 285 \tab 134 \tab 1048 \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 412 \tab 199 \tab 1241 \tab 8 \tab - \tab - \tab 1860 \cr
   1983 \tab 311 \tab 152 \tab 1131 \tab 5 \tab - \tab - \tab 1599 \cr
   1984 \tab 327 \tab 151 \tab 990 \tab 5 \tab - \tab - \tab 1473 \cr
   1985 \tab 339 \tab 157 \tab 1038 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 313 \tab 169 \tab 988 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 424 \tab 242 \tab 1149 \tab 4 \tab - \tab - \tab 1819 \cr
   1988 \tab 328 \tab 173 \tab 978 \tab 2 \tab - \tab - \tab 1481 \cr
   1989 \tab 327 \tab 202 \tab 989 \tab 19 \tab - \tab - \tab 1537 \cr
   1990 \tab 323 \tab 199 \tab 843 \tab 7 \tab - \tab - \tab 1372 \cr
   1991 \tab 374 \tab 205 \tab 937 \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab 373 \tab 216 \tab 1014 \tab 3 \tab - \tab - \tab 1606 \cr
   1994 \tab 752 \tab 393 \tab 1845 \tab 2 \tab - \tab - \tab 2992 \cr
   1996 \tab 743 \tab 463 \tab 1671 \tab 27 \tab - \tab - \tab 2904 \cr
   1998 \tab 777 \tab 456 \tab 1586 \tab 13 \tab - \tab - \tab 2832 \cr
   2000 \tab 741 \tab 720 \tab 1340 \tab 16 \tab - \tab - \tab 2817 \cr
   2002 \tab 883 \tab 668 \tab 1212 \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab 707 \tab 715 \tab 1390 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1226 \tab 1132 \tab 2150 \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab 527 \tab 528 \tab 967 \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 595 \tab 510 \tab 936 \tab 3 \tab - \tab - \tab 2044 \cr
   2012 \tab 533 \tab 502 \tab 936 \tab 3 \tab - \tab - \tab 1974 \cr
   2014 \tab 678 \tab 669 \tab 1185 \tab 6 \tab - \tab - \tab 2538 \cr
   2016 \tab 849 \tab 793 \tab 1223 \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab 677 \tab 666 \tab 1000 \tab 5 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab 2666 \tab 320 \tab 557 \tab - \tab - \tab 1 \tab 3544 \cr
   Total \tab 18137 \tab 11136 \tab 34346 \tab 137 \tab 4032 \tab 1 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away1}},
\code{\link{away2}},
\code{\link{gender1}},
\code{\link{mar1}},
\code{\link{mar2}},
\code{\link{old1}},
\code{\link{old2}},
\code{\link{relate2}},
\code{\link{where1}}
}
\concept{Household Members and Structure}
\keyword{variable}
