% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_584.R
\name{gender5}
\alias{gender5}
\title{Gender of 5th person}
\source{
General Social Survey https://gss.norc.org
}
\description{
gender5
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
2. Gender of fifth person
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[3]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5357/vshow}{GSS Data Explorer page for \code{gender5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab don't know \tab not available in this release \tab 3 \tab Total \cr
   1975 \tab 1202 \tab 143 \tab 145 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1369 \tab 70 \tab 60 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1264 \tab 130 \tab 135 \tab 1 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1293 \tab 120 \tab 118 \tab 1 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1283 \tab 90 \tab 93 \tab 2 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1610 \tab 128 \tab 120 \tab 2 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1376 \tab 107 \tab 112 \tab 4 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1280 \tab 98 \tab 95 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1347 \tab 99 \tab 88 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1265 \tab 101 \tab 103 \tab 1 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1577 \tab 128 \tab 113 \tab 1 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1328 \tab 66 \tab 86 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1356 \tab 89 \tab 88 \tab 4 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1237 \tab 55 \tab 78 \tab 2 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1367 \tab 74 \tab 75 \tab 1 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1451 \tab 75 \tab 77 \tab 3 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2729 \tab 140 \tab 120 \tab 3 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2637 \tab 127 \tab 131 \tab 9 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2588 \tab 115 \tab 125 \tab 4 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2559 \tab 122 \tab 133 \tab 3 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2556 \tab 107 \tab 102 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2567 \tab 113 \tab 130 \tab 2 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4086 \tab 210 \tab 206 \tab 1 \tab 7 \tab - \tab - \tab 4510 \cr
   2008 \tab 1826 \tab 99 \tab 90 \tab 2 \tab 6 \tab - \tab - \tab 2023 \cr
   2010 \tab 1861 \tab 76 \tab 101 \tab 4 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1775 \tab 101 \tab 93 \tab 3 \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab 2325 \tab 108 \tab 101 \tab 2 \tab 2 \tab - \tab - \tab 2538 \cr
   2016 \tab 2632 \tab 114 \tab 113 \tab 5 \tab 3 \tab - \tab - \tab 2867 \cr
   2018 \tab 2192 \tab 71 \tab 80 \tab 3 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab 3363 \tab 89 \tab 90 \tab - \tab - \tab - \tab 2 \tab 3544 \cr
   Total \tab 57301 \tab 3165 \tab 3201 \tab 64 \tab 24 \tab 4032 \tab 2 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away4}},
\code{\link{away5}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{old5}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{where4}},
\code{\link{where5}}
}
\concept{Household Members and Structure}
\keyword{variable}
