% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{gendereq}
\alias{gendereq}
\title{Gov resp to promote equality}
\source{
General Social Survey https://gss.norc.org
}
\description{
gendereq
}
\details{
Question (On the whole, do you think it should or should not be the government's responsibility to...)
Promote equality between men and women?  (Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6304/vshow}{GSS Data Explorer page for \code{gendereq}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
