% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{geneabrt}
\alias{geneabrt}
\title{Abort defective fetus}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneabrt
}
\details{
Question Suppose a test shows the baby has a serious genetic defect. Would you (yourself want to/ want your partner to) have an abortion if a test shows the baby has a serious genetic defect?
}
\section{Values}{

\itemize{
\item \verb{[1]} have abortion
\item \verb{[2]} no abortion
\item \verb{[7]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   A/C/- \tab 1990, 1996, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5089/vshow}{GSS Data Explorer page for \code{geneabrt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab have abortion \tab no abortion \tab refused \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab 455 \tab 377 \tab 347 \tab 20 \tab 165 \tab 8 \tab - \tab 1372 \cr
   1996 \tab 1936 \tab 404 \tab 434 \tab - \tab 49 \tab 81 \tab - \tab 2904 \cr
   2004 \tab 2 \tab 808 \tab 1643 \tab - \tab 312 \tab 47 \tab - \tab 2812 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 4741 \tab 1589 \tab 2424 \tab 20 \tab 526 \tab 136 \tab 0 \tab 9436 \cr
}
}

\seealso{
Other Genes: 
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{genegood}},
\code{\link{geneself}},
\code{\link{genetest}},
\code{\link{genetst1}}

Other Abortion: 
\code{\link{abany}},
\code{\link{abchoose}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abhave3}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{ablegal}},
\code{\link{abnomore}},
\code{\link{aborct}},
\code{\link{abpoor}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{abspno}},
\code{\link{salabort}}

Other Genetic Testing: 
\code{\link{cloning}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{geneexps}},
\code{\link{genegets}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{genegood}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{geneself}},
\code{\link{genetest}},
\code{\link{genetst1}},
\code{\link{genevig1}},
\code{\link{genevig2}},
\code{\link{tomatoes}}
}
\concept{Abortion}
\concept{Genes}
\concept{Genetic Testing}
\keyword{variable}
