% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{geneabrt2}
\alias{geneabrt2}
\title{R would want to abort fetus if there were genetic defects}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneabrt2
}
\details{
Question Suppose a test shows the baby has a serious genetic defect. Would you (yourself want/want the pregnant woman/want your partner) to have an abortion if a test shows the baby has a serious genetic defect?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7296/vshow}{GSS Data Explorer page for \code{geneabrt2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1168 \tab 335 \tab 644 \tab 149 \tab 52 \tab 2348 \cr
   Total \tab 1168 \tab 335 \tab 644 \tab 149 \tab 52 \tab 2348 \cr
}
}

\keyword{variable}
