% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{genecanx}
\alias{genecanx}
\title{Employer test cancer or safeworkplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
genecanx
}
\details{
Question A. Suppose workers will be exposed to some cancer-causing materials on the job. In that case, should employers have the right not to hire workers whose tests show they have an inherited tendency to develop certain forms of cancer, or should they be required to clean up the workplace so it is safe for everyone?
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1260/vshow}{GSS Data Explorer page for \code{genecanx}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not \tab don't know \tab no answer \tab Total \cr
   1991 \tab 761 \tab 45 \tab 667 \tab 34 \tab 10 \tab 1517 \cr
   Total \tab 761 \tab 45 \tab 667 \tab 34 \tab 10 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}
}
\concept{Work Organizations}
\keyword{variable}
