% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{genegets}
\alias{genegets}
\title{Employer vs. employee on access right to results}
\source{
General Social Survey https://gss.norc.org
}
\description{
genegets
}
\details{
Question C. Suppose a genetic screening test for certain forms of heart disease is made available by an employer to workers who want to take it. Who Do you think should have control over access to the test results -- the worker or the employer?
}
\section{Values}{

\itemize{
\item \verb{[1]} employer should
\item \verb{[2]} worker should
\item \verb{[3]} both - voluntary
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1262/vshow}{GSS Data Explorer page for \code{genegets}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab employer should \tab worker should \tab both - voluntary \tab don't know \tab no answer \tab Total \cr
   1991 \tab 524 \tab 83 \tab 715 \tab 157 \tab 31 \tab 7 \tab 1517 \cr
   Total \tab 524 \tab 83 \tab 715 \tab 157 \tab 31 \tab 7 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}
}
\concept{Work Organizations}
\keyword{variable}
