% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{genegood}
\alias{genegood}
\title{Genetic testing good or harm}
\source{
General Social Survey https://gss.norc.org
}
\description{
genegood
}
\details{
Question A. Some people say that genetic screening is a wonderful medical advance. Others think it may cause trouble. Based on what you know, do you think genetic screening will do more good than harm, or more harm than good?
}
\section{Values}{

\itemize{
\item \code{1} good ) harm
\item \code{2} harm ) good
\item \code{3} it depends
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5085/vshow}{GSS Data Explorer page for \code{genegood}}.
}

\seealso{
Other Genes: 
\code{\link{geneabrt}},
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{geneself}},
\code{\link{genetest}},
\code{\link{genetst1}}
}
\concept{Genes}
\keyword{variable}
