% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_134.R
\name{genehire}
\alias{genehire}
\title{Employers' right to hire by genetic tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
genehire
}
\details{
Question B. Should employers have the right not to hire workers if tests show they have an inherited tendency to develop certain forms of cancer or heart disease, or should they not have the right?
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1259/vshow}{GSS Data Explorer page for \code{genehire}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not \tab don't know \tab no answer \tab Total \cr
   1991 \tab 524 \tab 109 \tab 828 \tab 40 \tab 16 \tab 1517 \cr
   Total \tab 524 \tab 109 \tab 828 \tab 40 \tab 16 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{genejob}}
}
\concept{Work Organizations}
\keyword{variable}
