% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_134.R
\name{genejob}
\alias{genejob}
\title{Should employers give genetic tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
genejob
}
\details{
Question Here are some questions about a new scientific technology called "genetic screening."  Using genetic screening, it is now possible to tell whether someone has inherited a tendency to develop certain cancers and certain forms of heart disease. These tests do not mean that a person will always develop the disease, but only that he or she may do so, depending on other conditions.
A. Should employers have the right to give these genetic tests to people who are applying for a job, or shouldn't they have that right?
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1258/vshow}{GSS Data Explorer page for \code{genejob}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not \tab don't know \tab no answer \tab Total \cr
   1991 \tab - \tab 210 \tab 1220 \tab 70 \tab 17 \tab 1517 \cr
   Total \tab 0 \tab 210 \tab 1220 \tab 70 \tab 17 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{genehire}}
}
\concept{Work Organizations}
\keyword{variable}
