% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{genenvo1}
\alias{genenvo1}
\title{Gene or learning?}
\source{
General Social Survey https://gss.norc.org
}
\description{
genenvo1
}
\details{
Question (The description of the first person presented to the respondent - please check GENEVIG1 & GENEVIG2)(Please type in a number from 1 to 21):
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3032/vshow}{GSS Data Explorer page for \code{genenvo1}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 432 \tab 164 \tab 248 \tab 196 \tab 126 \tab 54 \tab 74 \tab 239 \tab 77 \tab 102 \tab 155 \tab 75 \tab 42 \tab 107 \tab 288 \tab 43 \tab 39 \tab 124 \tab 42 \tab 60 \tab 55 \tab 44 \tab 14 \tab 12 \tab 2812 \cr
   Total \tab 432 \tab 164 \tab 248 \tab 196 \tab 126 \tab 54 \tab 74 \tab 239 \tab 77 \tab 102 \tab 155 \tab 75 \tab 42 \tab 107 \tab 288 \tab 43 \tab 39 \tab 124 \tab 42 \tab 60 \tab 55 \tab 44 \tab 14 \tab 12 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{genenvo2}},
\code{\link{genevig1}},
\code{\link{genevig2}}
}
\concept{Genes and the Environment}
\keyword{variable}
