% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{genes}
\alias{genes}
\title{Traits are transferred over generation}
\source{
General Social Survey https://gss.norc.org
}
\description{
genes
}
\details{
Question Traits are transferred from generation to generation through the…
}
\section{Values}{

\itemize{
\item \verb{[1]} sperm only
\item \verb{[2]} egg only
\item \verb{[3]} sperm and egg
\item \verb{[4]} testes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3479/vshow}{GSS Data Explorer page for \code{genes}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab sperm only \tab egg only \tab sperm and egg \tab testes \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 95 \tab 31 \tab 1178 \tab 30 \tab 153 \tab 18 \tab 2023 \cr
   Total \tab 518 \tab 95 \tab 31 \tab 1178 \tab 30 \tab 153 \tab 18 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
