% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{genetst1}
\alias{genetst1}
\title{Know about genetic testing}
\source{
General Social Survey https://gss.norc.org
}
\description{
genetst1
}
\details{
Question B. How much would you say you have heard or read about genetic testing?
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} something but not very much
\item \code{3} nothing at all
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   A/C/- \tab 1996, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5084/vshow}{GSS Data Explorer page for \code{genetst1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a great deal \tab something but not very much \tab nothing at all \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2431 \tab 70 \tab 276 \tab 82 \tab 10 \tab 35 \tab 2904 \cr
   2004 \tab 2 \tab 508 \tab 1896 \tab 397 \tab 9 \tab - \tab 2812 \cr
   2018 \tab 1168 \tab 142 \tab 694 \tab 314 \tab 4 \tab 26 \tab 2348 \cr
   Total \tab 3601 \tab 720 \tab 2866 \tab 793 \tab 23 \tab 61 \tab 8064 \cr
}
}

\seealso{
Other Genes: 
\code{\link{genegood}},
\code{\link{genetest}}
}
\concept{Genes}
\keyword{variable}
