% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{genevig1}
\alias{genevig1}
\title{Vignette number used}
\source{
General Social Survey https://gss.norc.org
}
\description{
genevig1
}
\details{
Question Vignette set number given to respondent.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3030/vshow}{GSS Data Explorer page for \code{genevig1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab Total \cr
   2004 \tab - \tab 80 \tab 79 \tab 83 \tab 89 \tab 92 \tab 83 \tab 84 \tab 89 \tab 98 \tab 96 \tab 92 \tab 102 \tab 86 \tab 92 \tab 88 \tab 84 \tab 85 \tab 80 \tab 91 \tab 82 \tab 90 \tab 90 \tab 77 \tab 95 \tab 88 \tab 89 \tab 88 \tab 94 \tab 88 \tab 85 \tab 77 \tab 96 \tab 2812 \cr
   Total \tab 0 \tab 80 \tab 79 \tab 83 \tab 89 \tab 92 \tab 83 \tab 84 \tab 89 \tab 98 \tab 96 \tab 92 \tab 102 \tab 86 \tab 92 \tab 88 \tab 84 \tab 85 \tab 80 \tab 91 \tab 82 \tab 90 \tab 90 \tab 77 \tab 95 \tab 88 \tab 89 \tab 88 \tab 94 \tab 88 \tab 85 \tab 77 \tab 96 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genevig2}}
}
\concept{Genes and the Environment}
\keyword{variable}
