% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{genevig2}
\alias{genevig2}
\title{Vignette order used}
\source{
General Social Survey https://gss.norc.org
}
\description{
genevig2
}
\details{
Question Vignette order given to respondent.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3031/vshow}{GSS Data Explorer page for \code{genevig2}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab not available in this year \tab Total \cr
   2004 \tab - \tab 128 \tab 103 \tab 111 \tab 124 \tab 108 \tab 96 \tab 117 \tab 127 \tab 123 \tab 127 \tab 102 \tab 118 \tab 119 \tab 124 \tab 123 \tab 98 \tab 126 \tab 118 \tab 132 \tab 107 \tab 129 \tab 118 \tab 116 \tab 118 \tab - \tab 2812 \cr
   Total \tab 0 \tab 128 \tab 103 \tab 111 \tab 124 \tab 108 \tab 96 \tab 117 \tab 127 \tab 123 \tab 127 \tab 102 \tab 118 \tab 119 \tab 124 \tab 123 \tab 98 \tab 126 \tab 118 \tab 132 \tab 107 \tab 129 \tab 118 \tab 116 \tab 118 \tab 0 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{genevig1}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{owndoing}},
\code{\link{pessimst}},
\code{\link{sadblue}},
\code{\link{satself}},
\code{\link{selfish}},
\code{\link{slfrspct}}

Other Genetic Testing: 
\code{\link{cloning}},
\code{\link{geneabrt}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{geneexps}},
\code{\link{genegets}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{genegood}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{geneself}},
\code{\link{genetest}},
\code{\link{genetst1}},
\code{\link{genevig1}},
\code{\link{tomatoes}}
}
\concept{Genes and the Environment}
\concept{Genetic Testing}
\keyword{variable}
