% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{getaheadnv}
\alias{getaheadnv}
\title{Opinion of how people get ahead (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
getaheadnv
}
\details{
Question Some people say that people get ahead by their own hard work; others say that lucky breaks or help from other people are more important. Which do you think is most important?
}
\section{Values}{

\itemize{
\item \code{1} hard work most important
\item \code{2} hard work, luck equally important (phone mode only: volunteered
\item \code{3} luck most important
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7609/vshow}{GSS Data Explorer page for \code{getaheadnv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab hard work most important \tab hard work, luck equally important (phone mode only: volunteered \tab luck most important \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2714 \tab 979 \tab 33 \tab 300 \tab 3 \tab 3 \tab 4032 \cr
   2022 \tab 2981 \tab 422 \tab 4 \tab 134 \tab - \tab 3 \tab 3544 \cr
   Total \tab 5695 \tab 1401 \tab 37 \tab 434 \tab 3 \tab 6 \tab 7576 \cr
}
}

\keyword{variable}
