% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{getaid}
\alias{getaid}
\title{Ever received welfare?}
\source{
General Social Survey https://gss.norc.org
}
\description{
getaid
}
\details{
Question Have you personally ever received income from Aid to Families with Dependent Children (AFDC), General Assistance, Supplemental Security Income, or Food Stamps? CIRCLE ONE CODE ONLY.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/583/vshow}{GSS Data Explorer page for \code{getaid}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1986 \tab - \tab 281 \tab 1179 \tab 10 \tab 1470 \cr
   Total \tab 0 \tab 281 \tab 1179 \tab 10 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{getahead}},
\code{\link{govaid}},
\code{\link{parsol}},
\code{\link{unemp}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}}
}
\concept{Core}
\keyword{variable}
