% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{getlang}
\alias{getlang}
\title{Where did r learn other language}
\source{
General Social Survey https://gss.norc.org
}
\description{
getlang
}
\details{
Question Is that a language you first learned as a child at home, in school, or is it one that you learned elsewhere?
}
\section{Values}{

\itemize{
\item \code{1} childhood home
\item \code{2} school
\item \code{3} elsewhere
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2289/vshow}{GSS Data Explorer page for \code{getlang}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab childhood home \tab school \tab elsewhere \tab no answer \tab Total \cr
   2000 \tab 2450 \tab 183 \tab 116 \tab 58 \tab 10 \tab 2817 \cr
   2006 \tab 3837 \tab 342 \tab 230 \tab 99 \tab 2 \tab 4510 \cr
   Total \tab 6287 \tab 525 \tab 346 \tab 157 \tab 12 \tab 7327 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
