% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{gills}
\alias{gills}
\title{How fish get oxygen}
\source{
General Social Survey https://gss.norc.org
}
\description{
gills
}
\details{
Question How do most fish get the oxygen they need to survive?
}
\section{Values}{

\itemize{
\item \code{1} they take in water and break it down into hydrogen and oxygen
\item \code{2} using their gills, they take in oxygen that is dissolved in water
\item \code{3} they get their oxygen from the food they eat
\item \code{4} they come to the surface every few minutes to breathe air into their lungs
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3480/vshow}{GSS Data Explorer page for \code{gills}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
