% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{givchng}
\alias{givchng}
\title{Have given money to group advocating social change}
\source{
General Social Survey https://gss.norc.org
}
\description{
givchng
}
\details{
Question Over the past 5 years have you done any of the following to express your opinion about an issue or your support for a cause?
E. Give money to a group advocating social change
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2914/vshow}{GSS Data Explorer page for \code{givchng}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 326 \tab 1048 \tab 12 \tab 7 \tab 2765 \cr
   Total \tab 1372 \tab 326 \tab 1048 \tab 12 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{hlpneedy}},
\code{\link{partthon}}
}
\concept{Adult Transitions}
\keyword{variable}
