% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_323.R
\name{givchrty}
\alias{givchrty}
\title{R has given money to a charity}
\source{
General Social Survey https://gss.norc.org
}
\description{
givchrty
}
\details{
Question During the past 12 months, how often have you done each of the following things:
F. Given money to a charity
}
\section{Values}{

\itemize{
\item \verb{[1]} more than once a week
\item \verb{[2]} once a week
\item \verb{[3]} once a month
\item \verb{[4]} at least 2 or 3 times in the past year
\item \verb{[5]} once in the past year
\item \verb{[6]} not at all in the past year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2882/vshow}{GSS Data Explorer page for \code{givchrty}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab more than once a week \tab once a week \tab once a month \tab at least 2 or 3 times in the past year \tab once in the past year \tab not at all in the past year \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 32 \tab 108 \tab 258 \tab 465 \tab 197 \tab 298 \tab 10 \tab 4 \tab 2765 \cr
   2004 \tab 1472 \tab 46 \tab 113 \tab 250 \tab 438 \tab 205 \tab 282 \tab 4 \tab 2 \tab 2812 \cr
   2012 \tab 666 \tab 31 \tab 85 \tab 213 \tab 421 \tab 190 \tab 358 \tab 3 \tab 7 \tab 1974 \cr
   2014 \tab 1264 \tab 38 \tab 87 \tab 244 \tab 367 \tab 181 \tab 352 \tab 1 \tab 4 \tab 2538 \cr
   Total \tab 4795 \tab 147 \tab 393 \tab 965 \tab 1691 \tab 773 \tab 1290 \tab 18 \tab 17 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{cutahead}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{givblood}},
\code{\link{givhmlss}},
\code{\link{givseat}},
\code{\link{retchnge}},
\code{\link{volchrty}}
}
\concept{Altruism}
\keyword{variable}
