% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{givearts}
\alias{givearts}
\title{Contributions to arts culture humanities}
\source{
General Social Survey https://gss.norc.org
}
\description{
givearts
}
\details{
Question During the last year, approximately how much money did you and the other family members in your household contribute to each of the following:
D. Were any of your contributions to organizations involved in the arts, culture, or the humanities?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1142/vshow}{GSS Data Explorer page for \code{givearts}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 2022 \tab 307 \tab 405 \tab 3 \tab 95 \tab - \tab 2832 \cr
   Total \tab 2022 \tab 307 \tab 405 \tab 3 \tab 95 \tab 0 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{madenkid}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numbwfrd}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{othsch}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcand}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}

Other Contribute Money To: 
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}}

Other Voluntary Associations: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{bthgrp2a}},
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}},
\code{\link{bthgrp4a}},
\code{\link{bthgrp4b}},
\code{\link{bthgrp4c}},
\code{\link{bthgrp4d}},
\code{\link{bthgrp4e}},
\code{\link{bthgrp5a}},
\code{\link{bthgrp5b}},
\code{\link{bthgrp5c}},
\code{\link{bthgrp5d}},
\code{\link{bthgrp5e}},
\code{\link{caregrp}},
\code{\link{churchtx}},
\code{\link{churhgrp}},
\code{\link{frstgrp1}},
\code{\link{frstgrp2}},
\code{\link{frstgrp3}},
\code{\link{frstgrp4}},
\code{\link{frstgrp5}},
\code{\link{givart}},
\code{\link{givecong}},
\code{\link{giveduc}},
\code{\link{givegrp}},
\code{\link{givenvir}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{givfound}},
\code{\link{givhlth}},
\code{\link{givhuman}},
\code{\link{givinfrm}},
\code{\link{givintl}},
\code{\link{givoth}},
\code{\link{givpol}},
\code{\link{givpub}},
\code{\link{givrec}},
\code{\link{givrelig}},
\code{\link{givwork}},
\code{\link{givyouth}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grpparty}},
\code{\link{grppol}},
\code{\link{grprelig}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{grpwork}},
\code{\link{hrsart}},
\code{\link{hrseduc}},
\code{\link{hrsenvir}},
\code{\link{hrsfound}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrsinfrm}},
\code{\link{hrsintl}},
\code{\link{hrsoth}},
\code{\link{hrspol}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{hrswork}},
\code{\link{hrsyouth}},
\code{\link{infrmgrp}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{mempolit}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{monart}},
\code{\link{moneduc}},
\code{\link{monenvir}},
\code{\link{monfound}},
\code{\link{monhlth}},
\code{\link{monhuman}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}},
\code{\link{monpol}},
\code{\link{monpub}},
\code{\link{monrec}},
\code{\link{monrelig}},
\code{\link{monwork}},
\code{\link{monyouth}},
\code{\link{mostact}},
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{mtfrat4}},
\code{\link{mtgreek1}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mthobby1}},
\code{\link{mthobby2}},
\code{\link{mthobby3}},
\code{\link{mtinfrm1}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtlit3}},
\code{\link{mtlit4}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}},
\code{\link{mtother1}},
\code{\link{mtother2}},
\code{\link{mtother3}},
\code{\link{mtother4}},
\code{\link{mtother5}},
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}},
\code{\link{mtschl1}},
\code{\link{mtschl2}},
\code{\link{mtschl3}},
\code{\link{mtschl4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv7}},
\code{\link{mtserv8}},
\code{\link{mtsport1}},
\code{\link{mtsport2}},
\code{\link{mtsport3}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtunion4}},
\code{\link{mtvet1}},
\code{\link{mtvet2}},
\code{\link{mtvet3}},
\code{\link{mtvet4}},
\code{\link{mtvet5}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}},
\code{\link{mtyouth4}},
\code{\link{mtyouth5}},
\code{\link{numchurh}},
\code{\link{numfarm}},
\code{\link{numfrat}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numinfrm}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numpolit}},
\code{\link{numprof}},
\code{\link{numschl}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{partpol}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{totart}},
\code{\link{toteduc}},
\code{\link{totenvir}},
\code{\link{totfound}},
\code{\link{tothlth}},
\code{\link{tothuman}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totpub}},
\code{\link{totrec}},
\code{\link{totrelig}},
\code{\link{totwork}},
\code{\link{totyouth}},
\code{\link{typchurh}},
\code{\link{typfarm}},
\code{\link{typfrat}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typinfrm}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typother}},
\code{\link{typpolit}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typserv}},
\code{\link{typsport}},
\code{\link{typunion}},
\code{\link{typvet}},
\code{\link{typyouth}},
\code{\link{valart}},
\code{\link{valeduc}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhlth}},
\code{\link{valhuman}},
\code{\link{valinfrm}},
\code{\link{valintl}},
\code{\link{valoth}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}},
\code{\link{volacty2}},
\code{\link{volactyr}},
\code{\link{volart}},
\code{\link{volarts}},
\code{\link{voleduc}},
\code{\link{volenvir}},
\code{\link{volfound}},
\code{\link{volhlth}},
\code{\link{volhuman}},
\code{\link{volinfrm}},
\code{\link{volintl}},
\code{\link{volmonth}},
\code{\link{voloth}},
\code{\link{volpol}},
\code{\link{volpub}},
\code{\link{volrec}},
\code{\link{volrelig}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{volwork}},
\code{\link{volyouth}},
\code{\link{whoelse1}},
\code{\link{whoelse2}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}},
\code{\link{writegrp}},
\code{\link{yrchurh1}},
\code{\link{yrchurh2}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrfarm3}},
\code{\link{yrfrat1}},
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby1}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrnat1}},
\code{\link{yrnat2}},
\code{\link{yrnat3}},
\code{\link{yrother1}},
\code{\link{yrother2}},
\code{\link{yrother3}},
\code{\link{yrother4}},
\code{\link{yrother5}},
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}},
\code{\link{yrprof6}},
\code{\link{yrschl1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yrserv1}},
\code{\link{yrserv2}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}},
\code{\link{yrserv6}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrsport6}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{Contribute Money To}
\concept{Religion}
\concept{Voluntary Associations}
\keyword{variable}
