% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{giveoth}
\alias{giveoth}
\title{Given money non-religious charities}
\source{
General Social Survey https://gss.norc.org
}
\description{
giveoth
}
\details{
Question During the last year, approximately how much money did you and the other family members in your household contribute to each of the following:
C. To non-religious charities, organizations, or causes?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1141/vshow}{GSS Data Explorer page for \code{giveoth}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 1000 \tab 10000 \tab 110 \tab 119 \tab 12 \tab 120 \tab 1200 \tab 125 \tab 144 \tab 15 \tab 150 \tab 1500 \tab 15000 \tab 175 \tab 2 \tab 20 \tab 200 \tab 2000 \tab 225 \tab 24 \tab 25 \tab 250 \tab 2500 \tab 260 \tab 30 \tab 300 \tab 3000 \tab 30000 \tab 35 \tab 3500 \tab 40 \tab 400 \tab 45 \tab 450 \tab 47 \tab 480 \tab 5 \tab 50 \tab 500 \tab 5000 \tab 52 \tab 5500 \tab 56 \tab 60 \tab 600 \tab 6000 \tab 7 \tab 700 \tab 75 \tab 750 \tab 8 \tab 80 \tab 800 \tab 8000 \tab 8500 \tab don't know \tab Total \cr
   1998 \tab 1387 \tab 567 \tab 1 \tab 30 \tab 130 \tab 32 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 5 \tab 16 \tab 5 \tab 1 \tab 1 \tab 1 \tab 32 \tab 88 \tab 13 \tab 1 \tab 2 \tab 20 \tab 12 \tab 1 \tab 1 \tab 14 \tab 46 \tab 7 \tab 1 \tab 7 \tab 1 \tab 8 \tab 16 \tab 1 \tab 2 \tab 1 \tab 1 \tab 10 \tab 94 \tab 59 \tab 3 \tab 2 \tab 1 \tab 1 \tab 4 \tab 6 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab 161 \tab 2832 \cr
   Total \tab 1387 \tab 567 \tab 1 \tab 30 \tab 130 \tab 32 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 5 \tab 16 \tab 5 \tab 1 \tab 1 \tab 1 \tab 32 \tab 88 \tab 13 \tab 1 \tab 2 \tab 20 \tab 12 \tab 1 \tab 1 \tab 14 \tab 46 \tab 7 \tab 1 \tab 7 \tab 1 \tab 8 \tab 16 \tab 1 \tab 2 \tab 1 \tab 1 \tab 10 \tab 94 \tab 59 \tab 3 \tab 2 \tab 1 \tab 1 \tab 4 \tab 6 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab 161 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{givearts}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{kid5up}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}},
\code{\link{relhrs2}}
}
\concept{Religion}
\keyword{variable}
