% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{giverel}
\alias{giverel}
\title{Given money other religious org}
\source{
General Social Survey https://gss.norc.org
}
\description{
giverel
}
\details{
Question During the last year, approximately how much money did you and the other family members in your household contribute to each of the following:
B. To other religious organizations, programs, or causes?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1140/vshow}{GSS Data Explorer page for \code{giverel}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 1000 \tab 10000 \tab 104 \tab 12 \tab 1200 \tab 125 \tab 15 \tab 150 \tab 1500 \tab 1650 \tab 2 \tab 20 \tab 200 \tab 2000 \tab 240 \tab 25 \tab 250 \tab 2500 \tab 25000 \tab 30 \tab 300 \tab 3000 \tab 35 \tab 350 \tab 40 \tab 400 \tab 4000 \tab 450 \tab 4500 \tab 5 \tab 50 \tab 500 \tab 50000 \tab 60 \tab 600 \tab 700 \tab 72 \tab 75 \tab 8 \tab 800 \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 947 \tab 1 \tab 14 \tab 54 \tab 10 \tab 5 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 13 \tab 39 \tab 9 \tab 2 \tab 21 \tab 5 \tab 2 \tab 1 \tab 4 \tab 23 \tab 5 \tab 1 \tab 1 \tab 2 \tab 11 \tab 1 \tab 1 \tab 1 \tab 7 \tab 41 \tab 28 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 68 \tab 94 \tab 2832 \cr
   Total \tab 1387 \tab 947 \tab 1 \tab 14 \tab 54 \tab 10 \tab 5 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 13 \tab 39 \tab 9 \tab 2 \tab 21 \tab 5 \tab 2 \tab 1 \tab 4 \tab 23 \tab 5 \tab 1 \tab 1 \tab 2 \tab 11 \tab 1 \tab 1 \tab 1 \tab 7 \tab 41 \tab 28 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 68 \tab 94 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{givearts}},
\code{\link{giveoth}},
\code{\link{homesch}},
\code{\link{kid5up}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}},
\code{\link{relhrs2}}
}
\concept{Religion}
\keyword{variable}
