% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{givinffor}
\alias{givinffor}
\title{Should govt collect foreigner info without knowledge}
\source{
General Social Survey https://gss.norc.org
}
\description{
givinffor
}
\details{
Question (Some people think that governments should have the right to take certain measures in the name of national security. Others disagree.  Do you think that the American government should or should not have the right to do the following?)
Collect information about anyone living in other countries without their knowledge?  (Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should have the right
\item \verb{[2]} probably should have the right
\item \verb{[3]} probably should not have the right
\item \verb{[4]} definitely should not have the right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6319/vshow}{GSS Data Explorer page for \code{givinffor}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should have the right \tab probably should have the right \tab probably should not have the right \tab definitely should not have the right \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 224 \tab 372 \tab 449 \tab 304 \tab 28 \tab 13 \tab 2867 \cr
   Total \tab 1477 \tab 224 \tab 372 \tab 449 \tab 304 \tab 28 \tab 13 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{polsfair}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{wotrial}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
