% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{givinfusa}
\alias{givinfusa}
\title{Should govt collect citizen info without knowledge}
\source{
General Social Survey https://gss.norc.org
}
\description{
givinfusa
}
\details{
Question Some people think that governments should have the right to take certain measures in the name of national security. Others disagree.  Do you think that the American government should or should not have the right to do the following?
Collect information about anyone living in America without their knowledge?  Would you say...
}
\section{Values}{

\itemize{
\item \code{1} definitely should have the right
\item \code{2} probably should have the right
\item \code{3} probably should not have the right
\item \code{4} definitely should not have the right
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6318/vshow}{GSS Data Explorer page for \code{givinfusa}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{hlpinflu}},
\code{\link{polsfair}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{wotrial}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
