% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_471.R
\name{givlaw}
\alias{givlaw}
\title{How much ought lawyer make}
\source{
General Social Survey https://gss.norc.org
}
\description{
givlaw
}
\details{
Question Next, what do you think people in these jobs ought to be paid -- how much do you think they should earn each year  before taxes, regardless of what they actually get? Please write in how much they should earn each year.
L. A lawyer?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999992} more than now paid
\item \code{99999993} same as now paid
\item \code{99999994} less than now paid
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4290/vshow}{GSS Data Explorer page for \code{givlaw}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10000 \tab 105000 \tab 110000 \tab 120000 \tab 125000 \tab 130000 \tab 14000 \tab 149000 \tab 15000 \tab 150000 \tab 170000 \tab 175000 \tab 180000 \tab 1e+05 \tab $1 million (or more, for 1987 & 2000) \tab 2000 \tab 20000 \tab 2250 \tab 225000 \tab 24000 \tab 2500 \tab 25000 \tab 250000 \tab 26000 \tab 260000 \tab 27000 \tab 275000 \tab 28000 \tab 29000 \tab 2e+05 \tab 30000 \tab 31200 \tab 32000 \tab 35000 \tab 350000 \tab 38000 \tab 3e+05 \tab 40000 \tab 4200 \tab 420000 \tab 44000 \tab 45000 \tab 48000 \tab 49000 \tab 4e+05 \tab 50000 \tab 520 \tab 55000 \tab 5e+05 \tab 60000 \tab 65000 \tab 68000 \tab 70000 \tab 75000 \tab 750000 \tab 78000 \tab 80000 \tab 85000 \tab 90000 \tab 95000 \tab 9e+05 \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 4 \tab 1 \tab 3 \tab 11 \tab 16 \tab 2 \tab 2 \tab 1 \tab 4 \tab 94 \tab 2 \tab 8 \tab 1 \tab 210 \tab 8 \tab 1 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 65 \tab 17 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab 9 \tab 21 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 8 \tab 110 \tab 1 \tab 3 \tab 11 \tab 55 \tab 5 \tab 1 \tab 53 \tab 79 \tab 2 \tab 1 \tab 86 \tab 7 \tab 24 \tab 1 \tab 2 \tab 66 \tab 209 \tab 2817 \cr
   Total \tab 1545 \tab 4 \tab 1 \tab 3 \tab 11 \tab 16 \tab 2 \tab 2 \tab 1 \tab 4 \tab 94 \tab 2 \tab 8 \tab 1 \tab 210 \tab 8 \tab 1 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 65 \tab 17 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab 9 \tab 21 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 8 \tab 110 \tab 1 \tab 3 \tab 11 \tab 55 \tab 5 \tab 1 \tab 53 \tab 79 \tab 2 \tab 1 \tab 86 \tab 7 \tab 24 \tab 1 \tab 2 \tab 66 \tab 209 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givowner}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
