% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_470.R
\name{givmason}
\alias{givmason}
\title{How much ought mason make?}
\source{
General Social Survey https://gss.norc.org
}
\description{
givmason
}
\details{
Question Next, what do you think people in these jobs ought to be paid -- how much do you think they should earn each year  before taxes, regardless of what they actually get? Please write in how much they should earn each year.
A. First, about how much do you think a bricklayer should earn?
}
\section{Values}{

\itemize{
\item \verb{[1e+06]} $1 million (or more, for 1987 & 2000)
\item \verb{[99999992]} more than now paid
\item \verb{[99999993]} same as now paid
\item \verb{[99999994]} less than now paid
\item \verb{[99999997]} never worked
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4279/vshow}{GSS Data Explorer page for \code{givmason}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1000 \tab 10000 \tab 11000 \tab 11500 \tab 12000 \tab 13000 \tab 13440 \tab 14000 \tab 14560 \tab 15000 \tab 16000 \tab 17000 \tab 17500 \tab 1800 \tab 18000 \tab 19000 \tab 19200 \tab 1e+05 \tab $1 million (or more, for 1987 & 2000) \tab 20000 \tab 20800 \tab 21000 \tab 2200 \tab 22000 \tab 22880 \tab 23000 \tab 23040 \tab 23500 \tab 24000 \tab 24960 \tab 25000 \tab 26000 \tab 27000 \tab 27500 \tab 28000 \tab 29000 \tab 3000 \tab 30000 \tab 31000 \tab 31200 \tab 32000 \tab 33280 \tab 34000 \tab 35000 \tab 36000 \tab 37000 \tab 37500 \tab 38000 \tab 39000 \tab 39520 \tab 40000 \tab 44000 \tab 45000 \tab 450000 \tab 48000 \tab 5000 \tab 50000 \tab 5200 \tab 52000 \tab 55000 \tab 5e+05 \tab 6000 \tab 60000 \tab 62400 \tab 65000 \tab 7000 \tab 70000 \tab 75000 \tab 7999 \tab 8000 \tab 80000 \tab 8500 \tab 85000 \tab 89000 \tab 9000 \tab 90000 \tab more than now paid \tab same as now paid \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1987 \tab 255 \tab 1 \tab 17 \tab 2 \tab 1 \tab 14 \tab 3 \tab 1 \tab 2 \tab 1 \tab 62 \tab 11 \tab 9 \tab 1 \tab 1 \tab 33 \tab 4 \tab 1 \tab 2 \tab 1 \tab 171 \tab 2 \tab 3 \tab 1 \tab 14 \tab 2 \tab 5 \tab 1 \tab 1 \tab 12 \tab 1 \tab 200 \tab 5 \tab 6 \tab 2 \tab 15 \tab 5 \tab 2 \tab 276 \tab 2 \tab 4 \tab 5 \tab 1 \tab 3 \tab 95 \tab 7 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 109 \tab 1 \tab 17 \tab 1 \tab 1 \tab 3 \tab 58 \tab 1 \tab 4 \tab 3 \tab 1 \tab 3 \tab 8 \tab 1 \tab 2 \tab 2 \tab 2 \tab 6 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 76 \tab 231 \tab - \tab 1819 \cr
   Total \tab 255 \tab 1 \tab 17 \tab 2 \tab 1 \tab 14 \tab 3 \tab 1 \tab 2 \tab 1 \tab 62 \tab 11 \tab 9 \tab 1 \tab 1 \tab 33 \tab 4 \tab 1 \tab 2 \tab 1 \tab 171 \tab 2 \tab 3 \tab 1 \tab 14 \tab 2 \tab 5 \tab 1 \tab 1 \tab 12 \tab 1 \tab 200 \tab 5 \tab 6 \tab 2 \tab 15 \tab 5 \tab 2 \tab 276 \tab 2 \tab 4 \tab 5 \tab 1 \tab 3 \tab 95 \tab 7 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 109 \tab 1 \tab 17 \tab 1 \tab 1 \tab 3 \tab 58 \tab 1 \tab 4 \tab 3 \tab 1 \tab 3 \tab 8 \tab 1 \tab 2 \tab 2 \tab 2 \tab 6 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 76 \tab 231 \tab 0 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{books16}},
\code{\link{conage}},
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{goodlife}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{inequal7}},
\code{\link{justpay}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{occmobil}},
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oppol}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}},
\code{\link{opsex}},
\code{\link{opwlth}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{pasup}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychild}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payexec}},
\code{\link{payfam1}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payhard}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payresp}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{paysup}},
\code{\link{payunskl}},
\code{\link{rank}()},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}},
\code{\link{unsklrnk}}

Other Inequality, Income: 
\code{\link{eqincome}},
\code{\link{equalize}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{helppoor}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{inequal7}},
\code{\link{opoutcme}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{richpoor}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{ISSP Social Inequality}
\concept{Inequality, Income}
\concept{Occupation}
\keyword{variable}
