% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_205.R
\name{givpub}
\alias{givpub}
\title{Contributed money in pub-soc benefit}
\source{
General Social Survey https://gss.norc.org
}
\description{
givpub
}
\details{
Question Listed on this card are examples of many different fields in which people and families contribute money or other property for charitable purposes. I mean making a voluntary contribution and not with the intention of making a profit or obtaining goods and/or services for yourself. In which, if any, of the fields listed on this card have you and the members of your family or household contributed some money or other property in 1995? Just read off the letter of each field (DON'T READ. RECORD BELOW UNDER GIVHLTH-GIVOTH.)
F. Public/society benefit
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1935/vshow}{GSS Data Explorer page for \code{givpub}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 150 \tab 1134 \tab 160 \tab 2904 \cr
   Total \tab 1460 \tab 150 \tab 1134 \tab 160 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{givart}},
\code{\link{givenvir}},
\code{\link{givfound}},
\code{\link{givhuman}},
\code{\link{givpol}},
\code{\link{givrec}},
\code{\link{givrelig}},
\code{\link{givwork}},
\code{\link{givyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
