% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_323.R
\name{givseat}
\alias{givseat}
\title{R offered seat to a stranger during past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
givseat
}
\details{
Question During the past 12 months, how often have you done each of the following things:
G. Offered your seat on a bus or in a public place to a stranger who was standing
}
\section{Values}{

\itemize{
\item \code{1} more than once a week
\item \code{2} once a week
\item \code{3} once a month
\item \code{4} at least 2 or 3 times in the past year
\item \code{5} once in the past year
\item \code{6} not at all in the past year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2883/vshow}{GSS Data Explorer page for \code{givseat}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab more than once a week \tab once a week \tab once a month \tab at least 2 or 3 times in the past year \tab once in the past year \tab not at all in the past year \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 20 \tab 26 \tab 70 \tab 259 \tab 185 \tab 788 \tab 18 \tab 6 \tab 2765 \cr
   2004 \tab 1472 \tab 22 \tab 30 \tab 99 \tab 330 \tab 195 \tab 655 \tab 6 \tab 3 \tab 2812 \cr
   2012 \tab 666 \tab 24 \tab 24 \tab 75 \tab 297 \tab 186 \tab 686 \tab 6 \tab 10 \tab 1974 \cr
   2014 \tab 1264 \tab 27 \tab 23 \tab 87 \tab 240 \tab 196 \tab 687 \tab 9 \tab 5 \tab 2538 \cr
   Total \tab 4795 \tab 93 \tab 103 \tab 331 \tab 1126 \tab 762 \tab 2816 \tab 39 \tab 24 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{cutahead}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{givblood}},
\code{\link{givchrty}},
\code{\link{givhmlss}},
\code{\link{retchnge}},
\code{\link{volchrty}}
}
\concept{Altruism}
\keyword{variable}
