% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_471.R
\name{givunskl}
\alias{givunskl}
\title{How much ought unskilled worker make?}
\source{
General Social Survey https://gss.norc.org
}
\description{
givunskl
}
\details{
Question Next, what do you think people in these jobs ought to be paid -- how much do you think they should earn each year  before taxes, regardless of what they actually get? Please write in how much they should earn each year.
J. An unskilled worker in a factory?
}
\section{Values}{

\itemize{
\item \verb{[1e+06]} $1 million (or more, for 1987 & 2000)
\item \verb{[99999992]} more than now paid
\item \verb{[99999993]} same as now paid
\item \verb{[99999994]} less than now paid
\item \verb{[99999997]} never worked
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4288/vshow}{GSS Data Explorer page for \code{givunskl}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1000 \tab 10000 \tab 10400 \tab 10500 \tab 11000 \tab 11520 \tab 12000 \tab 12480 \tab 12500 \tab 13000 \tab 13500 \tab 14000 \tab 14400 \tab 14500 \tab 1500 \tab 15000 \tab 15600 \tab 16000 \tab 16500 \tab 17000 \tab 17500 \tab 18000 \tab 18500 \tab 19000 \tab 1e+05 \tab 2000 \tab 20000 \tab 20800 \tab 21000 \tab 22000 \tab 23000 \tab 24000 \tab 25000 \tab 26000 \tab 27000 \tab 2800 \tab 28000 \tab 29000 \tab 29999 \tab 2e+05 \tab 3000 \tab 30000 \tab 33000 \tab 3500 \tab 35000 \tab 39000 \tab 4000 \tab 40000 \tab 45000 \tab 450000 \tab 5000 \tab 50000 \tab 5200 \tab 52000 \tab 5500 \tab 6000 \tab 6968 \tab 7000 \tab 7500 \tab 7680 \tab 7800 \tab 8000 \tab 8320 \tab 8500 \tab 8989 \tab 9000 \tab 9600 \tab more than now paid \tab same as now paid \tab less than now paid \tab can't choose \tab no answer \tab 11440 \tab 125000 \tab 14560 \tab 18720 \tab 19200 \tab $1 million (or more, for 1987 & 2000) \tab 2250 \tab 2500 \tab 250000 \tab 31000 \tab 32000 \tab 33600 \tab 354000 \tab 38000 \tab 416 \tab 55000 \tab 58000 \tab 60000 \tab 75000 \tab 80000 \tab 10 \tab 1010 \tab 1100 \tab 1200 \tab 13347 \tab 15 \tab 150000 \tab 18 \tab 1e+07 \tab 20 \tab 200 \tab 21 \tab 23 \tab 25333 \tab 25900 \tab 25999 \tab 2800000 \tab 3 \tab 300 \tab 31200 \tab 320000 \tab 34000 \tab 35 \tab 350000 \tab 3500000 \tab 35360 \tab 35843 \tab 36000 \tab 36058 \tab 361847 \tab 37000 \tab 3e+05 \tab 3e+06 \tab 400 \tab 42000 \tab 4333 \tab 44000 \tab 440000 \tab 45 \tab 47000 \tab 48000 \tab 49000 \tab 49950 \tab 4e+05 \tab 5 \tab 52321 \tab 550 \tab 56000 \tab 560000 \tab 57000 \tab 58900 \tab 5e+05 \tab 5e+06 \tab 62000 \tab 65000 \tab 666666 \tab 68000 \tab 6e+06 \tab 70000 \tab 737362 \tab 75 \tab 7e+05 \tab 800 \tab 8179 \tab 85000 \tab 8e+06 \tab 900 \tab 90000 \tab 98876 \tab skipped on web \tab not available in this year \tab Total \cr
   1987 \tab 255 \tab 1 \tab 3 \tab 134 \tab 2 \tab 1 \tab 14 \tab 1 \tab 103 \tab 2 \tab 1 \tab 23 \tab 1 \tab 27 \tab 1 \tab 1 \tab 1 \tab 271 \tab 1 \tab 38 \tab 3 \tab 30 \tab 4 \tab 87 \tab 4 \tab 11 \tab 1 \tab 6 \tab 212 \tab 1 \tab 2 \tab 13 \tab 1 \tab 4 \tab 82 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 37 \tab 1 \tab 1 \tab 7 \tab 2 \tab 6 \tab 10 \tab 1 \tab 1 \tab 8 \tab 2 \tab 1 \tab 2 \tab 1 \tab 12 \tab 1 \tab 11 \tab 1 \tab 1 \tab 1 \tab 30 \tab 1 \tab 3 \tab 1 \tab 18 \tab 1 \tab 1 \tab 2 \tab 1 \tab 73 \tab 223 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   2000 \tab 1545 \tab - \tab 1 \tab 28 \tab - \tab - \tab - \tab - \tab 25 \tab - \tab - \tab 4 \tab - \tab 6 \tab - \tab - \tab - \tab 93 \tab - \tab 10 \tab - \tab 11 \tab - \tab 34 \tab - \tab 9 \tab 5 \tab 1 \tab 225 \tab - \tab - \tab 7 \tab 3 \tab 8 \tab 175 \tab - \tab 2 \tab - \tab 16 \tab - \tab - \tab 1 \tab 1 \tab 150 \tab 1 \tab - \tab 42 \tab - \tab - \tab 68 \tab 5 \tab - \tab 3 \tab 30 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 60 \tab 208 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 4 \tab 3 \tab 13 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 13 \tab - \tab 1 \tab - \tab 1 \tab - \tab 7 \tab - \tab 4 \tab 22 \tab 2 \tab 61 \tab 1 \tab 5 \tab 2 \tab 4 \tab 8 \tab 109 \tab 7 \tab 4 \tab - \tab 20 \tab 2 \tab - \tab 4 \tab 1 \tab 212 \tab 1 \tab - \tab 150 \tab 2 \tab 1 \tab 285 \tab 134 \tab 1 \tab 1 \tab 273 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 15 \tab 132 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 4 \tab 14 \tab - \tab - \tab 8 \tab - \tab 22 \tab - \tab 98 \tab 39 \tab 31 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 1 \tab 25 \tab 1 \tab 1 \tab 1 \tab 35 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 83 \tab - \tab 4032 \cr
   Total \tab 3829 \tab 5 \tab 7 \tab 175 \tab 2 \tab 1 \tab 14 \tab 1 \tab 131 \tab 2 \tab 1 \tab 29 \tab 1 \tab 33 \tab 1 \tab 1 \tab 3 \tab 377 \tab 1 \tab 49 \tab 3 \tab 42 \tab 4 \tab 128 \tab 4 \tab 24 \tab 28 \tab 9 \tab 498 \tab 2 \tab 7 \tab 22 \tab 8 \tab 20 \tab 366 \tab 11 \tab 7 \tab 1 \tab 37 \tab 4 \tab 1 \tab 6 \tab 7 \tab 399 \tab 3 \tab 1 \tab 199 \tab 4 \tab 7 \tab 363 \tab 140 \tab 2 \tab 12 \tab 305 \tab 1 \tab 4 \tab 1 \tab 12 \tab 1 \tab 17 \tab 1 \tab 1 \tab 1 \tab 34 \tab 1 \tab 3 \tab 1 \tab 20 \tab 1 \tab 1 \tab 2 \tab 1 \tab 148 \tab 563 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 6 \tab 5 \tab 18 \tab 1 \tab 1 \tab 9 \tab 1 \tab 23 \tab 1 \tab 104 \tab 42 \tab 32 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 1 \tab 25 \tab 1 \tab 1 \tab 1 \tab 35 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 83 \tab 0 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{books16}},
\code{\link{conage}},
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{goodlife}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{inequal7}},
\code{\link{justpay}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{occmobil}},
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oppol}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}},
\code{\link{opsex}},
\code{\link{opwlth}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{pasup}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychild}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payexec}},
\code{\link{payfam1}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payhard}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payresp}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{paysup}},
\code{\link{payunskl}},
\code{\link{rank}()},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}},
\code{\link{unsklrnk}}

Other Inequality, Income: 
\code{\link{eqincome}},
\code{\link{equalize}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{helppoor}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{inequal7}},
\code{\link{opoutcme}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{richpoor}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{ISSP Social Inequality}
\concept{Inequality, Income}
\concept{Occupation}
\keyword{variable}
