% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{godchnge}
\alias{godchnge}
\title{Describe your beliefs about god}
\source{
General Social Survey https://gss.norc.org
}
\description{
godchnge
}
\details{
Question Which best describes your beliefs about God?
}
\section{Values}{

\itemize{
\item \code{1} i don't believe in god now, and i never have
\item \code{2} i don't believe in god now, but i used to
\item \code{3} i believe in god now, but i didn't used to
\item \code{4} i believe in god now, and i always have
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4650/vshow}{GSS Data Explorer page for \code{godchnge}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
