% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_344.R
\name{godclose}
\alias{godclose}
\title{How close to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
godclose
}
\details{
Question In general, how close do you feel to God?
}
\section{Values}{

\itemize{
\item \code{1} not at all close
\item \code{2} somewhat close
\item \code{3} very close
\item \code{4} as close as possible
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3093/vshow}{GSS Data Explorer page for \code{godclose}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all close \tab somewhat close \tab very close \tab as close as possible \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 143 \tab 478 \tab 473 \tab 223 \tab 6 \tab 17 \tab 2812 \cr
   Total \tab 1472 \tab 143 \tab 478 \tab 473 \tab 223 \tab 6 \tab 17 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{blessngs}},
\code{\link{closrgod}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{joylifts}},
\code{\link{relsprt1}},
\code{\link{relsprt2}}
}
\concept{Spirituality and Religion}
\keyword{variable}
