% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{godlove}
\alias{godlove}
\title{Feel gods love directly or thru others}
\source{
General Social Survey https://gss.norc.org
}
\description{
godlove
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
E. I feel God's love for me, directly or through others.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never/almst nevr
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2112/vshow}{GSS Data Explorer page for \code{godlove}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never/almst nevr \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 237 \tab 372 \tab 246 \tab 210 \tab 182 \tab 143 \tab 21 \tab 34 \tab 2832 \cr
   Total \tab 1387 \tab 237 \tab 372 \tab 246 \tab 210 \tab 182 \tab 143 \tab 21 \tab 34 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{beausprt}},
\code{\link{effort}},
\code{\link{fidgety}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{uniongod}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
