% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{godmeans}
\alias{godmeans}
\title{Life meaningful because god exists}
\source{
General Social Survey https://gss.norc.org
}
\description{
godmeans
}
\details{
Question Do you agree or disagree with the following. . .
C. To me, life is meaningful only because God exists.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4659/vshow}{GSS Data Explorer page for \code{godmeans}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 298 \tab 305 \tab 312 \tab 190 \tab 133 \tab 58 \tab 63 \tab 1517 \cr
   1998 \tab 1548 \tab 242 \tab 282 \tab 285 \tab 240 \tab 120 \tab 50 \tab 65 \tab 2832 \cr
   2008 \tab 658 \tab 268 \tab 352 \tab 278 \tab 298 \tab 151 \tab 13 \tab 5 \tab 2023 \cr
   2018 \tab 1173 \tab 253 \tab 264 \tab 226 \tab 240 \tab 174 \tab 7 \tab 11 \tab 2348 \cr
   Total \tab 3537 \tab 1061 \tab 1203 \tab 1101 \tab 968 \tab 578 \tab 128 \tab 144 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
