% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{godsells}
\alias{godsells}
\title{Religious image to sell products}
\source{
General Social Survey https://gss.norc.org
}
\description{
godsells
}
\details{
Question Do you strongly approve of, somewhat approve of, neither approve of nor oppose, somewhat oppose, or strongly oppose the following public expressions of religion:
A. The use of religious "images" in public advertising to sell non-religious commercial products?
}
\section{Values}{

\itemize{
\item \code{1} strongly approve
\item \code{2} somewhat approve
\item \code{3} not apprve/oppse
\item \code{4} somewhat oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1137/vshow}{GSS Data Explorer page for \code{godsells}}.
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsport}},
\code{\link{numbwfrd}},
\code{\link{numfrend}}
}
\concept{Religion}
\keyword{variable}
