% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{godswill}
\alias{godswill}
\title{How likely x s situation caused by god s will}
\source{
General Social Survey https://gss.norc.org
}
\description{
godswill
}
\details{
Question In your opinion, how likely is it that NAME's situation might be caused by (READ a-f) -- very likely, somewhat likely, not very likely, or not at all likely?
F. God's will
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1993/vshow}{GSS Data Explorer page for \code{godswill}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 67 \tab 134 \tab 334 \tab 827 \tab 66 \tab 16 \tab 2904 \cr
   2006 \tab 3073 \tab 55 \tab 126 \tab 332 \tab 868 \tab 42 \tab 14 \tab 4510 \cr
   2018 \tab 1175 \tab 47 \tab 107 \tab 303 \tab 664 \tab 20 \tab 32 \tab 2348 \cr
   Total \tab 5708 \tab 169 \tab 367 \tab 969 \tab 2359 \tab 128 \tab 62 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
