% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{goodlife}
\alias{goodlife}
\title{Standard of living of r will improve}
\source{
General Social Survey https://gss.norc.org
}
\description{
goodlife
}
\details{
Question The way things are in America, people like me and my family have a good chance of improving our standard of living -- do you agree or disagree?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/- \tab full \cr
   1994 \tab B/C \tab partial \cr
   1996 \tab B/C \tab partial \cr
   1998 \tab B/C \tab partial \cr
   2000 \tab B/C \tab partial \cr
   2002 \tab B/C \tab partial \cr
   2004 \tab B/C \tab partial \cr
   2006 \tab B/C \tab partial \cr
   2008 \tab B/C \tab partial \cr
   2010 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab B/C \tab partial \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4254/vshow}{GSS Data Explorer page for \code{goodlife}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{books16}},
\code{\link{conage}},
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{inequal7}},
\code{\link{justpay}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{occmobil}},
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oppol}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}},
\code{\link{opsex}},
\code{\link{opwlth}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{pasup}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychild}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payexec}},
\code{\link{payfam1}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payhard}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payresp}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{paysup}},
\code{\link{payunskl}},
\code{\link{rank}()},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
