% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_286.R
\name{googlesn}
\alias{googlesn}
\title{Use googlesn}
\source{
General Social Survey https://gss.norc.org
}
\description{
googlesn
}
\details{
Question (Which of the following social networking or social media sites are you a member or regular user of?)
Google+ -  Google's social network platform.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6276/vshow}{GSS Data Explorer page for \code{googlesn}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 1483 \tab 476 \tab 897 \tab 11 \tab 2867 \cr
   Total \tab 1483 \tab 476 \tab 897 \tab 11 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{facebook}},
\code{\link{instagrm}},
\code{\link{intrecnt}},
\code{\link{linkedin}},
\code{\link{snapchat}},
\code{\link{tumblr}},
\code{\link{twitter}},
\code{\link{whatsapp}}
}
\concept{Social Media Usage}
\keyword{variable}
