% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{goownway}
\alias{goownway}
\title{Importance of following own conscience}
\source{
General Social Survey https://gss.norc.org
}
\description{
goownway
}
\details{
Question People have many differing views about what makes a person a good Christian or Jew. Please tell me how important is each of the following to you. Once again, please indicate where you would place your feelings on a scale from 1 to 5. How important is it …
D. To follow one's conscience even if it means going against what the churches or synagogues say and do.
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[5]} not very important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1092/vshow}{GSS Data Explorer page for \code{goownway}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab 2 \tab 3 \tab 4 \tab not very important \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 605 \tab 333 \tab 303 \tab 88 \tab 110 \tab 24 \tab 18 \tab 1481 \cr
   1998 \tab 1387 \tab 576 \tab 234 \tab 313 \tab 92 \tab 137 \tab 78 \tab 15 \tab 2832 \cr
   Total \tab 1387 \tab 1181 \tab 567 \tab 616 \tab 180 \tab 247 \tab 102 \tab 33 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
