% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{govaid}
\alias{govaid}
\title{Ever receive welfare, unemp insur, etc.}
\source{
General Social Survey https://gss.norc.org
}
\description{
govaid
}
\details{
Question Did you ever--because of sickness, unemployment, or any other reason--receive anything like welfare, unemployment insurance, or other aid from government agencies?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/582/vshow}{GSS Data Explorer page for \code{govaid}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 502 \tab 1001 \tab 1 \tab - \tab 1504 \cr
   1974 \tab - \tab 481 \tab 999 \tab 2 \tab 2 \tab 1484 \cr
   1975 \tab - \tab 511 \tab 978 \tab - \tab 1 \tab 1490 \cr
   1976 \tab - \tab 518 \tab 980 \tab - \tab 1 \tab 1499 \cr
   1977 \tab - \tab 582 \tab 947 \tab 1 \tab - \tab 1530 \cr
   1978 \tab - \tab 570 \tab 959 \tab 3 \tab - \tab 1532 \cr
   1980 \tab - \tab 547 \tab 916 \tab 5 \tab - \tab 1468 \cr
   1983 \tab - \tab 614 \tab 980 \tab 3 \tab 2 \tab 1599 \cr
   Total \tab 0 \tab 4325 \tab 7760 \tab 15 \tab 6 \tab 12106 \cr
}
}

\seealso{
Other Core: 
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{parsol}},
\code{\link{unemp}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}}
}
\concept{Core}
\keyword{variable}
