% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{govcare}
\alias{govcare}
\title{Govt concern for social welfare}
\source{
General Social Survey https://gss.norc.org
}
\description{
govcare
}
\details{
Question B. For some time there have been discussions about how concerned the government should be about the social welfare of people. We have listed here two different opinions. With which opinion do you agree the most?  READ STATEMENTS A & B. CIRCLE ONE CODE.
}
\section{Values}{

\itemize{
\item \code{1} too much care
\item \code{2} on right track
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/242/vshow}{GSS Data Explorer page for \code{govcare}}.
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}},
\code{\link{tax}}
}
\concept{Core}
\concept{Economy}
\keyword{variable}
