% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{goveqinc1}
\alias{goveqinc1}
\title{Govnmnt should fix income differences}
\source{
General Social Survey https://gss.norc.org
}
\description{
goveqinc1
}
\details{
Question To what extent do you agree or disagree with the following statements?  It is the responsibility of the government to reduce the differences in income between people with high incomes and those with low incomes.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7480/vshow}{GSS Data Explorer page for \code{goveqinc1}}.
}

\keyword{variable}
