% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{governor}
\alias{governor}
\title{Did r correctly name governor}
\source{
General Social Survey https://gss.norc.org
}
\description{
governor
}
\details{
Question We would want to know how well known the different governmental leaders are around here? Could you tell me the name of the governor of this state?
}
\section{Values}{

\itemize{
\item \verb{[1]} correct
\item \verb{[2]} incorrect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1035/vshow}{GSS Data Explorer page for \code{governor}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab correct \tab incorrect \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 1372 \tab 82 \tab 350 \tab 15 \tab 1819 \cr
   Total \tab 0 \tab 1372 \tab 82 \tab 350 \tab 15 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
