% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{govfnaid}
\alias{govfnaid}
\title{Govt should give financial assistance to college students from low-income famili}
\source{
General Social Survey https://gss.norc.org
}
\description{
govfnaid
}
\details{
Question The government should give financial assistance to college students from low-income families, even if it might require a tax increase to pay for it
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7554/vshow}{GSS Data Explorer page for \code{govfnaid}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2250 \tab 281 \tab 681 \tab 419 \tab 242 \tab 90 \tab 52 \tab 2 \tab 15 \tab 4032 \cr
   Total \tab 2250 \tab 281 \tab 681 \tab 419 \tab 242 \tab 90 \tab 52 \tab 2 \tab 15 \tab 4032 \cr
}
}

\keyword{variable}
