% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{govineq2}
\alias{govineq2}
\title{Us govt success in reducing inc differences}
\source{
General Social Survey https://gss.norc.org
}
\description{
govineq2
}
\details{
Question How successful do you think the government in America is nowadays in reducing the differences in income between people with high incomes and people with low incomes?
}
\section{Values}{

\itemize{
\item \code{1} very successful
\item \code{2} quite successful
\item \code{3} neither successful nor unsuccessful
\item \code{4} quite unsuccessful
\item \code{5} very unsuccessful
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7508/vshow}{GSS Data Explorer page for \code{govineq2}}.
}

\keyword{variable}
