% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_463.R
\name{govlazy}
\alias{govlazy}
\title{The social benefits from the government make people lazy}
\source{
General Social Survey https://gss.norc.org
}
\description{
govlazy
}
\details{
Question The social benefits from the government make people lazy.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7240/vshow}{GSS Data Explorer page for \code{govlazy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 171 \tab 419 \tab 251 \tab 226 \tab 90 \tab 15 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 171 \tab 419 \tab 251 \tab 226 \tab 90 \tab 15 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
