% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_472.R
\name{govless}
\alias{govless}
\title{Govmnt should spend less on poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
govless
}
\details{
Question Do you agree or disagree?
E. The government should spend less on benefits for the poor.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4299/vshow}{GSS Data Explorer page for \code{govless}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 52 \tab 201 \tab 306 \tab 663 \tab 274 \tab 33 \tab 35 \tab 1819 \cr
   Total \tab 255 \tab 52 \tab 201 \tab 306 \tab 663 \tab 274 \tab 33 \tab 35 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givrocc}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incgap}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
