% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{govmentl}
\alias{govmentl}
\title{Should govt provide health care for mental illness}
\source{
General Social Survey https://gss.norc.org
}
\description{
govmentl
}
\details{
Question On the whole, do you think it should or should not be the government''s responsibility to provide mental health care for persons with mental illnesses?
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{8} cant choose
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2039/vshow}{GSS Data Explorer page for \code{govmentl}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 2196 \tab 243 \tab 274 \tab 86 \tab 53 \tab 40 \tab 12 \tab 2904 \cr
   2006 \tab 3808 \tab 266 \tab 293 \tab 104 \tab 16 \tab 14 \tab 9 \tab 4510 \cr
   Total \tab 6004 \tab 509 \tab 567 \tab 190 \tab 69 \tab 54 \tab 21 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mntlas1}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}},
\code{\link{spmentl}}
}
\concept{Mental Health}
\keyword{variable}
