% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{govngos}
\alias{govngos}
\title{Decision left to natl govt representative or citizen or}
\source{
General Social Survey https://gss.norc.org
}
\description{
govngos
}
\details{
Question Which of these two statements comes closer to your view?
}
\section{Values}{

\itemize{
\item \verb{[1]} should be left to govt representatives
\item \verb{[2]} citizen orgs should be involved
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4950/vshow}{GSS Data Explorer page for \code{govngos}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should be left to govt representatives \tab citizen orgs should be involved \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 516 \tab 812 \tab 126 \tab 18 \tab 2812 \cr
   Total \tab 1340 \tab 516 \tab 812 \tab 126 \tab 18 \tab 2812 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
