% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{govt30}
\alias{govt30}
\title{R use govt site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
govt30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
F. Government information?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} 3-5 times
\item \code{4} more than 5 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2399/vshow}{GSS Data Explorer page for \code{govt30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 360 \tab 191 \tab 58 \tab 59 \tab 74 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 501 \tab 248 \tab 108 \tab 132 \tab 94 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 341 \tab 171 \tab 75 \tab 91 \tab 12 \tab - \tab 2812 \cr
   Total \tab 5878 \tab 1202 \tab 610 \tab 241 \tab 282 \tab 180 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{bookmark}},
\code{\link{catdrcty}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{fin30}},
\code{\link{hyperlnk}},
\code{\link{news30}},
\code{\link{schl30}},
\code{\link{work30}}
}
\concept{Information Society}
\keyword{variable}
