% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{govtbiz}
\alias{govtbiz}
\title{Government do more for environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
govtbiz
}
\details{
Question B. And which of these two groups do you think is making more effort to look after the environment:
}
\section{Values}{

\itemize{
\item \verb{[1]} government
\item \verb{[2]} or business and industry?
\item \verb{[3]} both equally
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4785/vshow}{GSS Data Explorer page for \code{govtbiz}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab government \tab or business and industry? \tab both equally \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 492 \tab 136 \tab 303 \tab 241 \tab 104 \tab 2817 \cr
   Total \tab 1541 \tab 492 \tab 136 \tab 303 \tab 241 \tab 104 \tab 2817 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
