% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{govtpow}
\alias{govtpow}
\title{Power of federal govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
govtpow
}
\details{
Question And what about the federal government, does it have too much power or too little power?
}
\section{Values}{

\itemize{
\item \verb{[1]} far too much power
\item \verb{[2]} too much power
\item \verb{[3]} about the right amount of power
\item \verb{[4]} too little power
\item \verb{[5]} far too little power
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1996 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4070/vshow}{GSS Data Explorer page for \code{govtpow}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab far too much power \tab too much power \tab about the right amount of power \tab too little power \tab far too little power \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1985 \tab 857 \tab 104 \tab 261 \tab 240 \tab 25 \tab 2 \tab 35 \tab 10 \tab - \tab 1534 \cr
   1990 \tab 155 \tab 145 \tab 404 \tab 520 \tab 48 \tab 3 \tab 73 \tab 24 \tab - \tab 1372 \cr
   1996 \tab 1572 \tab 308 \tab 508 \tab 368 \tab 40 \tab 6 \tab 54 \tab 48 \tab - \tab 2904 \cr
   Total \tab 2584 \tab 557 \tab 1173 \tab 1128 \tab 113 \tab 11 \tab 162 \tab 82 \tab 0 \tab 5810 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidneedy}},
\code{\link{aidold}},
\code{\link{aidsmart}},
\code{\link{aidunemp}},
\code{\link{beltup}},
\code{\link{bribe}},
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{buspow}},
\code{\link{bustax}},
\code{\link{cctv}},
\code{\link{colop}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{cutdebt}},
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{databank}},
\code{\link{demworks}},
\code{\link{emonitor}},
\code{\link{eqincome}},
\code{\link{equalize}},
\code{\link{exploit}},
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{fewtrsty}},
\code{\link{gendereq}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{govtinfo}},
\code{\link{grnlaws}},
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}},
\code{\link{hlphitec}},
\code{\link{hlpinflu}},
\code{\link{hlthcare}},
\code{\link{hsbasics}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{jobsall}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidout}},
\code{\link{kidskips}},
\code{\link{kidxfilm}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{mustret}},
\code{\link{nosmoke}},
\code{\link{obeylaw}},
\code{\link{opfamily}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{polgbeco}},
\code{\link{polint}},
\code{\link{polsfair}},
\code{\link{prespop}},
\code{\link{pricecon}},
\code{\link{progtax}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{protstrs}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{prvdschl}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{racists}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{ratetone}},
\code{\link{revoltrs}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}},
\code{\link{unionsok}},
\code{\link{verdict}},
\code{\link{volacty2}},
\code{\link{volactyr}},
\code{\link{wotrial}},
\code{\link{youinflu}}

Other Federal Government: 
\code{\link{advfront}},
\code{\link{affrmact}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{concong}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{congeth}},
\code{\link{congovt}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{demworks}},
\code{\link{equalize}},
\code{\link{fedtrust}},
\code{\link{govcheat}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hlthcare}},
\code{\link{hspgovt}},
\code{\link{immcosts}},
\code{\link{immfare}},
\code{\link{jobaff}},
\code{\link{jobsall}},
\code{\link{knomentl}},
\code{\link{leadsci}},
\code{\link{mustret}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdiscm}},
\code{\link{natdisin}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natrecon}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{owneth}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{pricecon}},
\code{\link{reqinfo}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{scresrch}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secterr}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{spmentl}},
\code{\link{spnatdis}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{taxcheat}},
\code{\link{taxspend}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{whtgovt}}

Other Political: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{anomia7}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{bribe}},
\code{\link{buspow}},
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf11}},
\code{\link{caninf12}},
\code{\link{caninf13}},
\code{\link{caninf14}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf6}},
\code{\link{caninf7}},
\code{\link{caninf8}},
\code{\link{caninf9}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{civic}},
\code{\link{cntctgov}},
\code{\link{congeth}},
\code{\link{contemp}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{corruptn}},
\code{\link{demworks}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{engballt}},
\code{\link{ethorgs}},
\code{\link{fixmistk}},
\code{\link{gavepol}},
\code{\link{giveoth}},
\code{\link{govdook}},
\code{\link{governor}},
\code{\link{govngos}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{interpol}},
\code{\link{intpol}},
\code{\link{joindem}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{leftrght}},
\code{\link{libtemp}},
\code{\link{localgvt}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{ntcitvte}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{owneth}},
\code{\link{partpol}},
\code{\link{partyid}},
\code{\link{polactve}},
\code{\link{polcamp}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff2}},
\code{\link{poleff20}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy3}},
\code{\link{polfrel}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polinfo}},
\code{\link{polint}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polnews}},
\code{\link{polpapr}},
\code{\link{polrally}},
\code{\link{polsfair}},
\code{\link{poltv}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{polwww}},
\code{\link{powerun}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prespop}},
\code{\link{refrndms}},
\code{\link{relcand}},
\code{\link{schlhead}},
\code{\link{servepeo}},
\code{\link{signdpet}},
\code{\link{swayvote}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usrep}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{workpol}}
}
\concept{Federal Government}
\concept{ISSP Role of Gov't}
\concept{Political}
\keyword{variable}
