% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{grassnv}
\alias{grassnv}
\title{Should marijuana be made legal (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
grassnv
}
\details{
Question Do you think the use of marijuana should be made legal or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} should be legal
\item \verb{[2]} should not be legal
\item \verb{[3]} no opinion (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7615/vshow}{GSS Data Explorer page for \code{grassnv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab should be legal \tab should not be legal \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2719 \tab 959 \tab 336 \tab 8 \tab 1 \tab 9 \tab 4032 \cr
   2022 \tab 2951 \tab 449 \tab 140 \tab - \tab 1 \tab 3 \tab 3544 \cr
   2024 \tab 2659 \tab 453 \tab 191 \tab 1 \tab 2 \tab 3 \tab 3309 \cr
   Total \tab 8329 \tab 1861 \tab 667 \tab 9 \tab 4 \tab 15 \tab 10885 \cr
}
}

\keyword{variable}
