% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{grkidnum}
\alias{grkidnum}
\title{Number of adult grandchildren}
\source{
General Social Survey https://gss.norc.org
}
\description{
grkidnum
}
\details{
Question Now thinking of all your adult relatives -- those still living and age 18 or older.
How many of each do you have?  READ EACH ITEM. CODE A NUMBER OR ZERO FOR EACH.
B. Adult Grandchildren
}
\section{Values}{

\itemize{
\item \verb{[96]} 96 or more
\item \verb{[97]} # dk but ) 0
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4154/vshow}{GSS Data Explorer page for \code{grkidnum}}.

Counts by year:\tabular{lllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 2 \tab 20 \tab 23 \tab 3 \tab 31 \tab 4 \tab 5 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab # dk but ) 0 \tab don't know \tab no answer \tab Total \cr
   1986 \tab - \tab 1249 \tab 55 \tab 3 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 39 \tab 1 \tab 1 \tab 23 \tab 1 \tab 13 \tab 16 \tab 14 \tab 1 \tab 7 \tab 10 \tab 4 \tab 7 \tab 3 \tab 9 \tab 1470 \cr
   Total \tab 0 \tab 1249 \tab 55 \tab 3 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 39 \tab 1 \tab 1 \tab 23 \tab 1 \tab 13 \tab 16 \tab 14 \tab 1 \tab 7 \tab 10 \tab 4 \tab 7 \tab 3 \tab 9 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
