% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{grndemo}
\alias{grndemo}
\title{Protested for envir issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
grndemo
}
\details{
Question In the last five years, have you…
C. ...taken part in a protest or demonstration about an environmental issue?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, i have
\item \verb{[2]} no, i have not
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4776/vshow}{GSS Data Explorer page for \code{grndemo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, i have \tab no, i have not \tab no answer \tab can't choose \tab skipped on web \tab not available in this year \tab Total \cr
   1993 \tab 49 \tab 40 \tab 1390 \tab 127 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 39 \tab 1250 \tab 95 \tab 2 \tab - \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 37 \tab 1123 \tab 115 \tab 1 \tab - \tab - \tab 2817 \cr
   2010 \tab 614 \tab 31 \tab 1390 \tab 7 \tab 2 \tab - \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 89 \tab 1728 \tab 196 \tab - \tab 16 \tab - \tab 4032 \cr
   Total \tab 5813 \tab 236 \tab 6881 \tab 540 \tab 5 \tab 16 \tab 0 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{anrights}},
\code{\link{antests}},
\code{\link{bizpeop}},
\code{\link{busdecid}},
\code{\link{busgrn}},
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{chemfam}},
\code{\link{chemfree}},
\code{\link{chemgen}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{econgrn}},
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{excldimm}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grncon}},
\code{\link{grnecon}},
\code{\link{grneffme}},
\code{\link{grnexagg}},
\code{\link{grngroup}},
\code{\link{grnintl}},
\code{\link{grnmoney}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grnsign}},
\code{\link{grnsol}},
\code{\link{grntaxes}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{h2oless}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{helpharm}},
\code{\link{ihlpgrn}},
\code{\link{impgrn}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{ldcgrn}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{nobuygrn}},
\code{\link{nomeat}},
\code{\link{nukeacc}},
\code{\link{nukefam}},
\code{\link{nukegen}},
\code{\link{obeythnk}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{peopgrn}},
\code{\link{popgrwth}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{pubdecid}},
\code{\link{recycle}},
\code{\link{redcehme}},
\code{\link{resnatur}},
\code{\link{scifaith}},
\code{\link{scigrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{sciworse}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{tempgen1}},
\code{\link{toodifme}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}},
\code{\link{usdoenuf}},
\code{\link{waterfam}},
\code{\link{watergen}}

Other Environment: 
\code{\link{amprogrn}},
\code{\link{anrights}},
\code{\link{anscitst}},
\code{\link{antests}},
\code{\link{artseals}},
\code{\link{bizbstgw}},
\code{\link{bizbstnk}},
\code{\link{bizinfgw}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{bizpeop}},
\code{\link{busdecid}},
\code{\link{busgrn}},
\code{\link{caremost}},
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{chemfam}},
\code{\link{chemfree}},
\code{\link{chemgen}},
\code{\link{comorsci}},
\code{\link{drivless}},
\code{\link{econgrn}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{extinct}},
\code{\link{gasregs}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grncon}},
\code{\link{grnecon}},
\code{\link{grneffme}},
\code{\link{grnexagg}},
\code{\link{grngroup}},
\code{\link{grnintl}},
\code{\link{grnlaws}},
\code{\link{grnmoney}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grnsign}},
\code{\link{grnsol}},
\code{\link{grntaxes}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{h2oless}},
\code{\link{harmsgrn}},
\code{\link{helpharm}},
\code{\link{ihlpgrn}},
\code{\link{impgrn}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{intenvir}},
\code{\link{inuitway}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{knwgw}},
\code{\link{ldcgrn}},
\code{\link{morempg}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natrecon}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{nobuygrn}},
\code{\link{noicecap}},
\code{\link{nomeat}},
\code{\link{nukeacc}},
\code{\link{nukeelec}},
\code{\link{nukefam}},
\code{\link{nukegen}},
\code{\link{othssame}},
\code{\link{penguins}},
\code{\link{peopgovt}},
\code{\link{peopgrn}},
\code{\link{polbstgw}},
\code{\link{polbstnk}},
\code{\link{polinfgw}},
\code{\link{polinfnk}},
\code{\link{polnuke}},
\code{\link{popgrwth}},
\code{\link{pubdecid}},
\code{\link{recycle}},
\code{\link{redcehme}},
\code{\link{resnatur}},
\code{\link{sciagrgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{sciworse}},
\code{\link{sealevel}},
\code{\link{spenviro}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{tempgen1}},
\code{\link{toodifme}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{usdoenuf}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
