% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{grnecon}
\alias{grnecon}
\title{Worry too much about envir, too little econ}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnecon
}
\details{
Question And please check one box for each of these statements to show how much you agree or disagree with it.
A. We worry too much about the future of the environment, and not enough about prices and jobs today.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4725/vshow}{GSS Data Explorer page for \code{grnecon}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 163 \tab 454 \tab 223 \tab 502 \tab 145 \tab 19 \tab 51 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 139 \tab 422 \tab 184 \tab 447 \tab 143 \tab 15 \tab 36 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 127 \tab 328 \tab 213 \tab 398 \tab 123 \tab 20 \tab 67 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 151 \tab 447 \tab 224 \tab 459 \tab 114 \tab 30 \tab 5 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 114 \tab 310 \tab 453 \tab 574 \tab 344 \tab 56 \tab 167 \tab 11 \tab 4032 \cr
   Total \tab 5813 \tab 694 \tab 1961 \tab 1297 \tab 2380 \tab 869 \tab 140 \tab 326 \tab 11 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{anrights}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scifaith}},
\code{\link{scigrn}},
\code{\link{sciworse}}
}
\concept{ISSP Environment}
\keyword{variable}
