% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{grnexagg}
\alias{grnexagg}
\title{Environmental threats exaggerated}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnexagg
}
\details{
Question How much do you agree or disagree with each of these statements?
C. Many of the claims about environmental threats are exaggerated.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4781/vshow}{GSS Data Explorer page for \code{grnexagg}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}

Other Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
