% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{grnintl}
\alias{grnintl}
\title{Intl agreements for environment problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnintl
}
\details{
Question How much do you agree or disagree with each of these statements?
A. For environmental problems, there should be international agreements that American and other countries should be made to follow.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4787/vshow}{GSS Data Explorer page for \code{grnintl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 278 \tab 619 \tab 166 \tab 60 \tab 9 \tab 73 \tab 71 \tab 2817 \cr
   2010 \tab 614 \tab 354 \tab 711 \tab 184 \tab 58 \tab 33 \tab 83 \tab 7 \tab 2044 \cr
   Total \tab 2155 \tab 632 \tab 1330 \tab 350 \tab 118 \tab 42 \tab 156 \tab 78 \tab 4861 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
