% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_520.R
\name{grnprog}
\alias{grnprog}
\title{Worry too much about progress harming envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnprog
}
\details{
Question How much do you agree or disagree with each of the following statements.
A. People worry too much about human progress harming the environment.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4729/vshow}{GSS Data Explorer page for \code{grnprog}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 68 \tab 416 \tab 259 \tab 602 \tab 127 \tab 27 \tab 58 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 63 \tab 425 \tab 204 \tab 537 \tab 96 \tab 28 \tab 33 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 56 \tab 275 \tab 260 \tab 457 \tab 103 \tab 36 \tab 89 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 63 \tab 467 \tab 262 \tab 506 \tab 72 \tab 55 \tab 5 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 72 \tab 312 \tab 459 \tab 658 \tab 271 \tab 66 \tab 171 \tab 20 \tab 4032 \cr
   Total \tab 5813 \tab 322 \tab 1895 \tab 1444 \tab 2760 \tab 669 \tab 212 \tab 356 \tab 20 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grntaxes}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{resnatur}}
}
\concept{ISSP Environment}
\keyword{variable}
