% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{grntest1}
\alias{grntest1}
\title{All will die if exposed to radioact}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntest1
}
\details{
Question And for each of these statements, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?.
A. If someone is exposed to any amount of radioactivity, they are certain to die as a result.
}
\section{Values}{

\itemize{
\item \code{1} definitely true
\item \code{2} probably true
\item \code{3} probably not true
\item \code{4} definitely not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4746/vshow}{GSS Data Explorer page for \code{grntest1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 77 \tab 302 \tab 591 \tab 419 \tab 87 \tab 81 \tab 1606 \cr
   1994 \tab 1606 \tab 66 \tab 300 \tab 527 \tab 390 \tab 62 \tab 41 \tab 2992 \cr
   2000 \tab 1541 \tab 79 \tab 250 \tab 424 \tab 338 \tab 119 \tab 66 \tab 2817 \cr
   Total \tab 3196 \tab 222 \tab 852 \tab 1542 \tab 1147 \tab 268 \tab 188 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{toodifme}}
}
\concept{ISSP Environment}
\keyword{variable}
