% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{grntest2}
\alias{grntest2}
\title{Nuke waste dangerous for thousands of yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntest2
}
\details{
Question And for each of these statements, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
B. Some radioactive waste from nuclear power stations will stay dangerous for thousands of years.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4747/vshow}{GSS Data Explorer page for \code{grntest2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 490 \tab 781 \tab 102 \tab 20 \tab 83 \tab 81 \tab 1606 \cr
   1994 \tab 1606 \tab 431 \tab 730 \tab 103 \tab 10 \tab 72 \tab 40 \tab 2992 \cr
   Total \tab 1655 \tab 921 \tab 1511 \tab 205 \tab 30 \tab 155 \tab 121 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{toodifme}}
}
\concept{ISSP Environment}
\keyword{variable}
