% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{grntest3}
\alias{grntest3}
\title{Greenhouse effect caused by hole in atmosphere}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntest3
}
\details{
Question And for each of these statements, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
C. The greenhouse effect is caused by a hole in the earth's atmosphere.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4748/vshow}{GSS Data Explorer page for \code{grntest3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 267 \tab 609 \tab 188 \tab 187 \tab 222 \tab 84 \tab 1606 \cr
   1994 \tab 1606 \tab 211 \tab 548 \tab 211 \tab 143 \tab 228 \tab 45 \tab 2992 \cr
   2000 \tab 1541 \tab 202 \tab 443 \tab 221 \tab 149 \tab 192 \tab 69 \tab 2817 \cr
   Total \tab 3196 \tab 680 \tab 1600 \tab 620 \tab 479 \tab 642 \tab 198 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
