% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{grntest4}
\alias{grntest4}
\title{Using coal or gas contrib to greenhouse effect}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntest4
}
\details{
Question And for each of these statements, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
D. Every time we use coal or oil or gas, we contribute to the greenhouse effect.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4749/vshow}{GSS Data Explorer page for \code{grntest4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 264 \tab 680 \tab 287 \tab 51 \tab 191 \tab 84 \tab 1606 \cr
   1994 \tab 1606 \tab 186 \tab 629 \tab 286 \tab 57 \tab 185 \tab 43 \tab 2992 \cr
   2000 \tab 1541 \tab 215 \tab 527 \tab 230 \tab 56 \tab 184 \tab 64 \tab 2817 \cr
   Total \tab 3196 \tab 665 \tab 1836 \tab 803 \tab 164 \tab 560 \tab 191 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
