% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{grntest6}
\alias{grntest6}
\title{Humans are main cause of plant & animal extinct}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntest6
}
\details{
Question And for each of these statements, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
F. Human beings are the main cause of plant and animal species dying out.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4751/vshow}{GSS Data Explorer page for \code{grntest6}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 233 \tab 692 \tab 321 \tab 157 \tab 72 \tab 82 \tab 1606 \cr
   1994 \tab 1606 \tab 196 \tab 584 \tab 349 \tab 142 \tab 73 \tab 42 \tab 2992 \cr
   Total \tab 1655 \tab 429 \tab 1276 \tab 670 \tab 299 \tab 145 \tab 124 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
