% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_365.R
\name{grpboth1}
\alias{grpboth1}
\title{Both belong to the same group on the list}
\source{
General Social Survey https://gss.norc.org
}
\description{
grpboth1
}
\details{
Question A. Do you and first name both belong to one or more of the same groups on the list?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3298/vshow}{GSS Data Explorer page for \code{grpboth1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2653 \tab 86 \tab 20 \tab 1 \tab 52 \tab 2812 \cr
   Total \tab 2653 \tab 86 \tab 20 \tab 1 \tab 52 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
