% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_365.R
\name{grpboth2}
\alias{grpboth2}
\title{Both belong to the same group on the list}
\source{
General Social Survey https://gss.norc.org
}
\description{
grpboth2
}
\details{
Question B. Do you and second name both belong to one or more of the same groups on the list?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3299/vshow}{GSS Data Explorer page for \code{grpboth2}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet1}},
\code{\link{grpboth1}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
