% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{grpother}
\alias{grpother}
\title{R belongs to another voluntary association}
\source{
General Social Survey https://gss.norc.org
}
\description{
grpother
}
\details{
Question People sometimes belong to different kinds of groups or associations. For each type of group, please indicate whether you:
-Belong and actively participate,
-Belong but don't actively participate,
-Used to belong but do not any more,
-Have never belonged to it.
E. Another voluntary association
}
\section{Values}{

\itemize{
\item \verb{[1]} belong and actively participate
\item \verb{[2]} belong but dont participate
\item \verb{[3]} used to belong
\item \verb{[4]} never belonged
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4932/vshow}{GSS Data Explorer page for \code{grpother}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab belong and actively participate \tab belong but dont participate \tab used to belong \tab never belonged \tab no answer \tab can't choose \tab not available in this year \tab Total \cr
   2004 \tab 1340 \tab 301 \tab 87 \tab 347 \tab 733 \tab 4 \tab - \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 270 \tab 100 \tab 313 \tab 547 \tab 3 \tab 31 \tab - \tab 2538 \cr
   Total \tab 2614 \tab 571 \tab 187 \tab 660 \tab 1280 \tab 7 \tab 31 \tab 0 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{buypol}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{citviews}},
\code{\link{cntctgov}},
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{eqtreat}},
\code{\link{fixmistk}},
\code{\link{govdook}},
\code{\link{govngos}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{gvtrghts}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{hlthall}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{leftrght}},
\code{\link{leftrght1}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ntcitvte}},
\code{\link{obeylaws}},
\code{\link{oppsegov}},
\code{\link{othreasn}},
\code{\link{owndecns}},
\code{\link{paytaxes}},
\code{\link{polactve}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polnews}},
\code{\link{polopts}},
\code{\link{poltalk}},
\code{\link{powerun}},
\code{\link{racmeet}},
\code{\link{refrndms}},
\code{\link{relmeet}},
\code{\link{revmeet}},
\code{\link{rghtsmin}},
\code{\link{servepeo}},
\code{\link{servpub}},
\code{\link{signdpet}},
\code{\link{solok}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usmca1}},
\code{\link{usmca2}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Social Networks: 
\code{\link{acqasian}},
\code{\link{acqattnd}},
\code{\link{acqblack}},
\code{\link{acqbrnda}},
\code{\link{acqchild}},
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqdems}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhme}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmno}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqfmwht}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjans}},
\code{\link{acqjose}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqlinda}},
\code{\link{acqmaria}},
\code{\link{acqmark}},
\code{\link{acqmils}},
\code{\link{acqmyrac}},
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhlib}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}},
\code{\link{acqnhwht}},
\code{\link{acqnoatt}},
\code{\link{acqprisn}},
\code{\link{acqrachl}},
\code{\link{acqreps}},
\code{\link{acqshawn}},
\code{\link{acqsocs}},
\code{\link{acqunemp}},
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvalib}},
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhme}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkno}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}},
\code{\link{acqwkwht}},
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{comamind}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{fricall}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndknow}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{grpsprts}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{hlpinflu}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{lentto}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{peocntct}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkedto}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}},
\code{\link{trtasian}},
\code{\link{trtattnd}},
\code{\link{trtblack}},
\code{\link{trtbrnda}},
\code{\link{trtchild}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtcops}},
\code{\link{trtdems}},
\code{\link{trtelecs}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtjose}},
\code{\link{trtkaren}},
\code{\link{trtkeith}},
\code{\link{trtkevin}},
\code{\link{trtlaws}},
\code{\link{trtlib}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtmils}},
\code{\link{trtmyrac}},
\code{\link{trtnoatt}},
\code{\link{trtprisn}},
\code{\link{trtrachl}},
\code{\link{trtreps}},
\code{\link{trtshawn}},
\code{\link{trtsocs}},
\code{\link{trtunemp}},
\code{\link{trtwhite}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{youinflu}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{ISSP Citizenship}
\concept{Social Networks}
\keyword{variable}
