% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{grppol}
\alias{grppol}
\title{Participated in activity of political party}
\source{
General Social Survey https://gss.norc.org
}
\description{
grppol
}
\details{
Question People sometimes belong to different kinds of groups or associations. The next few questions contain different types of groups. For each type of group, please select a response to say whether you have participated in the activities of this group in the past 12 months.
A. A political party, club or association
}
\section{Values}{

\itemize{
\item \code{1} i have participated more than twice
\item \code{2} i have participated once or twice
\item \code{3} i belong to such a group but never participate
\item \code{4} i do not belong to such a group
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4207/vshow}{GSS Data Explorer page for \code{grppol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab i have participated more than twice \tab i have participated once or twice \tab i belong to such a group but never participate \tab i do not belong to such a group \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 113 \tab 100 \tab 75 \tab 853 \tab 1 \tab 7 \tab 2765 \cr
   Total \tab 1616 \tab 113 \tab 100 \tab 75 \tab 853 \tab 1 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
