% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{grprelig}
\alias{grprelig}
\title{R belongs to a church or othr religious organization}
\source{
General Social Survey https://gss.norc.org
}
\description{
grprelig
}
\details{
Question People sometimes belong to different kinds of groups or associations. For each type of group, please indicate whether you:
-Belong and actively participate,
-Belong but don't actively participate,
-Used to belong but do not any more,
-Have never belonged to it.
C. A church or other religious organization
}
\section{Values}{

\itemize{
\item \verb{[1]} belong and actively participate
\item \verb{[2]} belong but dont participate
\item \verb{[3]} used to belong
\item \verb{[4]} never belonged
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4930/vshow}{GSS Data Explorer page for \code{grprelig}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab belong and actively participate \tab belong but dont participate \tab used to belong \tab never belonged \tab no answer \tab can't choose \tab not available in this year \tab Total \cr
   2004 \tab 1340 \tab 588 \tab 315 \tab 316 \tab 246 \tab 7 \tab - \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 424 \tab 256 \tab 316 \tab 244 \tab 6 \tab 18 \tab - \tab 2538 \cr
   Total \tab 2614 \tab 1012 \tab 571 \tab 632 \tab 490 \tab 13 \tab 18 \tab 0 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{buypol}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{citviews}},
\code{\link{cntctgov}},
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{eqtreat}},
\code{\link{fixmistk}},
\code{\link{govdook}},
\code{\link{govngos}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{gvtrghts}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{hlthall}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{leftrght}},
\code{\link{leftrght1}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ntcitvte}},
\code{\link{obeylaws}},
\code{\link{oppsegov}},
\code{\link{othreasn}},
\code{\link{owndecns}},
\code{\link{paytaxes}},
\code{\link{polactve}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polnews}},
\code{\link{polopts}},
\code{\link{poltalk}},
\code{\link{powerun}},
\code{\link{racmeet}},
\code{\link{refrndms}},
\code{\link{relmeet}},
\code{\link{revmeet}},
\code{\link{rghtsmin}},
\code{\link{servepeo}},
\code{\link{servpub}},
\code{\link{signdpet}},
\code{\link{solok}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usmca1}},
\code{\link{usmca2}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Voluntary Associations: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{bthgrp2a}},
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}},
\code{\link{bthgrp4a}},
\code{\link{bthgrp4b}},
\code{\link{bthgrp4c}},
\code{\link{bthgrp4d}},
\code{\link{bthgrp4e}},
\code{\link{bthgrp5a}},
\code{\link{bthgrp5b}},
\code{\link{bthgrp5c}},
\code{\link{bthgrp5d}},
\code{\link{bthgrp5e}},
\code{\link{caregrp}},
\code{\link{churchtx}},
\code{\link{churhgrp}},
\code{\link{frstgrp1}},
\code{\link{frstgrp2}},
\code{\link{frstgrp3}},
\code{\link{frstgrp4}},
\code{\link{frstgrp5}},
\code{\link{givart}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveduc}},
\code{\link{givegrp}},
\code{\link{givenvir}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{givfound}},
\code{\link{givhlth}},
\code{\link{givhuman}},
\code{\link{givinfrm}},
\code{\link{givintl}},
\code{\link{givoth}},
\code{\link{givpol}},
\code{\link{givpub}},
\code{\link{givrec}},
\code{\link{givrelig}},
\code{\link{givwork}},
\code{\link{givyouth}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grpparty}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{grpwork}},
\code{\link{hrsart}},
\code{\link{hrseduc}},
\code{\link{hrsenvir}},
\code{\link{hrsfound}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrsinfrm}},
\code{\link{hrsintl}},
\code{\link{hrsoth}},
\code{\link{hrspol}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{hrswork}},
\code{\link{hrsyouth}},
\code{\link{infrmgrp}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{mempolit}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{monart}},
\code{\link{moneduc}},
\code{\link{monenvir}},
\code{\link{monfound}},
\code{\link{monhlth}},
\code{\link{monhuman}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}},
\code{\link{monpol}},
\code{\link{monpub}},
\code{\link{monrec}},
\code{\link{monrelig}},
\code{\link{monwork}},
\code{\link{monyouth}},
\code{\link{mostact}},
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{mtfrat4}},
\code{\link{mtgreek1}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mthobby1}},
\code{\link{mthobby2}},
\code{\link{mthobby3}},
\code{\link{mtinfrm1}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtlit3}},
\code{\link{mtlit4}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}},
\code{\link{mtother1}},
\code{\link{mtother2}},
\code{\link{mtother3}},
\code{\link{mtother4}},
\code{\link{mtother5}},
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}},
\code{\link{mtschl1}},
\code{\link{mtschl2}},
\code{\link{mtschl3}},
\code{\link{mtschl4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv7}},
\code{\link{mtserv8}},
\code{\link{mtsport1}},
\code{\link{mtsport2}},
\code{\link{mtsport3}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtunion4}},
\code{\link{mtvet1}},
\code{\link{mtvet2}},
\code{\link{mtvet3}},
\code{\link{mtvet4}},
\code{\link{mtvet5}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}},
\code{\link{mtyouth4}},
\code{\link{mtyouth5}},
\code{\link{numchurh}},
\code{\link{numfarm}},
\code{\link{numfrat}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numinfrm}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numpolit}},
\code{\link{numprof}},
\code{\link{numschl}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{partpol}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{totart}},
\code{\link{toteduc}},
\code{\link{totenvir}},
\code{\link{totfound}},
\code{\link{tothlth}},
\code{\link{tothuman}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totpub}},
\code{\link{totrec}},
\code{\link{totrelig}},
\code{\link{totwork}},
\code{\link{totyouth}},
\code{\link{typchurh}},
\code{\link{typfarm}},
\code{\link{typfrat}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typinfrm}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typother}},
\code{\link{typpolit}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typserv}},
\code{\link{typsport}},
\code{\link{typunion}},
\code{\link{typvet}},
\code{\link{typyouth}},
\code{\link{valart}},
\code{\link{valeduc}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhlth}},
\code{\link{valhuman}},
\code{\link{valinfrm}},
\code{\link{valintl}},
\code{\link{valoth}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}},
\code{\link{volacty2}},
\code{\link{volactyr}},
\code{\link{volart}},
\code{\link{volarts}},
\code{\link{voleduc}},
\code{\link{volenvir}},
\code{\link{volfound}},
\code{\link{volhlth}},
\code{\link{volhuman}},
\code{\link{volinfrm}},
\code{\link{volintl}},
\code{\link{volmonth}},
\code{\link{voloth}},
\code{\link{volpol}},
\code{\link{volpub}},
\code{\link{volrec}},
\code{\link{volrelig}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{volwork}},
\code{\link{volyouth}},
\code{\link{whoelse1}},
\code{\link{whoelse2}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}},
\code{\link{writegrp}},
\code{\link{yrchurh1}},
\code{\link{yrchurh2}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrfarm3}},
\code{\link{yrfrat1}},
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby1}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrnat1}},
\code{\link{yrnat2}},
\code{\link{yrnat3}},
\code{\link{yrother1}},
\code{\link{yrother2}},
\code{\link{yrother3}},
\code{\link{yrother4}},
\code{\link{yrother5}},
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}},
\code{\link{yrprof6}},
\code{\link{yrschl1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yrserv1}},
\code{\link{yrserv2}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}},
\code{\link{yrserv6}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrsport6}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{ISSP Citizenship}
\concept{Voluntary Associations}
\keyword{variable}
