% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{grpunion}
\alias{grpunion}
\title{Participated in activity of a trade union}
\source{
General Social Survey https://gss.norc.org
}
\description{
grpunion
}
\details{
Question People sometimes belong to different kinds of groups or associations. The next few questions contain different types of groups. For each type of group, please select a response to say whether you have participated in the activities of this group in the past 12 months.
B. A trade union or professional association
}
\section{Values}{

\itemize{
\item \verb{[1]} i have participated more than twice
\item \verb{[2]} i have participated once or twice
\item \verb{[3]} i belong to such a group but never participate
\item \verb{[4]} i do not belong to such a group
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4208/vshow}{GSS Data Explorer page for \code{grpunion}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab i have participated more than twice \tab i have participated once or twice \tab i belong to such a group but never participate \tab i do not belong to such a group \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 101 \tab 86 \tab 90 \tab 864 \tab 1 \tab 7 \tab 2765 \cr
   Total \tab 1616 \tab 101 \tab 86 \tab 90 \tab 864 \tab 1 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
