% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{grpwork}
\alias{grpwork}
\title{R belongs to a trade union or professional association}
\source{
General Social Survey https://gss.norc.org
}
\description{
grpwork
}
\details{
Question People sometimes belong to different kinds of groups or associations. For each type of group, please indicate whether you:
-Belong and actively participate,
-Belong but don’t actively participate,
-Used to belong but do not any more,
-Have never belonged to it.
B. A trade union, business, or professional association
}
\section{Values}{

\itemize{
\item \code{1} belong and actively participate
\item \code{2} belong but dont participate
\item \code{3} used to belong
\item \code{4} never belonged
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4929/vshow}{GSS Data Explorer page for \code{grpwork}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab belong and actively participate \tab belong but dont participate \tab used to belong \tab never belonged \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 173 \tab 137 \tab 280 \tab 877 \tab 1 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 115 \tab 129 \tab 224 \tab 775 \tab 18 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 288 \tab 266 \tab 504 \tab 1652 \tab 19 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{polopts}},
\code{\link{rghtsmin}},
\code{\link{solok}}
}
\concept{ISSP Citizenship}
\keyword{variable}
