% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssrdoc_funs.R
\name{gss_which_ballots}
\alias{gss_which_ballots}
\title{Which ballots did a particular question appear on in each GSS year?}
\usage{
gss_which_ballots(question)
}
\arguments{
\item{question}{A GSS variable name (in lower case)

Columns returned are
\itemize{
\item \code{variable} Variable name
\item \code{year} GSS year
\item \code{ballots} Ballots, separated by a slash \code{/}, with \code{-} for blank or inapplicable
\item \code{availability} Can be \code{full} or \code{partial}
}}
}
\value{
A tibble with the name of the variable and the ballot information.
}
\description{
Which ballots did a particular question appear on in each GSS year?
}
\examples{
\dontrun{
gss_which_ballots(fefam)
}

}
