% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssrdoc-package.R
\docType{package}
\name{gssrdoc-package}
\alias{gssrdoc}
\alias{gssrdoc-package}
\title{gssrdoc: Document General Social Survey Variables}
\description{
The General Social Survey (GSS) is a long-running, mostly annual survey of US households. It is administered by the National Opinion Research Center (NORC). This package contains the a tibble with information on the survey variables, together with every variable documented as an R help page. For more information on the GSS see \url{http://gss.norc.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kjhealy/gssrdoc}
  \item Report bugs at \url{https://github.com/kjhealy/gssrdoc/issues}
}

}
\author{
\strong{Maintainer}: Kieran Healy \email{kieran.healy@duke.edu} (\href{https://orcid.org/0000-0001-9114-981X}{ORCID})

}
\keyword{internal}
