% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{gun}
\alias{gun}
\title{Ever threatened with gun or shot at}
\source{
General Social Survey https://gss.norc.org
}
\description{
gun
}
\details{
Question Have you ever been threatened with a gun, or shot at?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/661/vshow}{GSS Data Explorer page for \code{gun}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 246 \tab 1256 \tab 2 \tab - \tab 1504 \cr
   1975 \tab - \tab 257 \tab 1229 \tab 3 \tab 1 \tab 1490 \cr
   1976 \tab - \tab 256 \tab 1241 \tab 1 \tab 1 \tab 1499 \cr
   1978 \tab - \tab 311 \tab 1220 \tab - \tab 1 \tab 1532 \cr
   1980 \tab - \tab 308 \tab 1156 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 326 \tab 1270 \tab 1 \tab 2 \tab 1599 \cr
   1984 \tab - \tab 290 \tab 1180 \tab 2 \tab 1 \tab 1473 \cr
   1986 \tab - \tab 296 \tab 1171 \tab - \tab 3 \tab 1470 \cr
   1987 \tab - \tab 384 \tab 1430 \tab 5 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 217 \tab 779 \tab - \tab 1 \tab 1481 \cr
   1989 \tab 502 \tab 200 \tab 832 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 178 \tab 719 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 201 \tab 810 \tab 6 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 201 \tab 855 \tab 1 \tab - \tab 1606 \cr
   1994 \tab 2481 \tab 104 \tab 405 \tab 1 \tab 1 \tab 2992 \cr
   Total \tab 4989 \tab 3775 \tab 15553 \tab 30 \tab 12 \tab 24359 \cr
}
}

\seealso{
Other Core: 
\code{\link{antiwar}},
\code{\link{civright}},
\code{\link{gunage}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitnum}},
\code{\link{prowar}},
\code{\link{school}},
\code{\link{strike}}
}
\concept{Core}
\keyword{variable}
