% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{gunage}
\alias{gunage}
\title{Threatened or shot at as child or adult}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunage
}
\details{
Question Have you ever been threatened with a gun, or shot at?
A. Did this happen to you as a child or as an adult?
}
\section{Values}{

\itemize{
\item \code{1} child
\item \code{2} adult
\item \code{3} both
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/662/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab adult \tab both \tab child \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1258 \tab 207 \tab 6 \tab 29 \tab 1 \tab 3 \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1230 \tab 214 \tab 7 \tab 32 \tab - \tab 7 \tab - \tab 1490 \cr
   1976 \tab 1242 \tab 206 \tab 12 \tab 37 \tab 1 \tab 1 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1221 \tab 249 \tab 17 \tab 44 \tab 1 \tab - \tab - \tab 1532 \cr
   1980 \tab 1157 \tab 257 \tab 13 \tab 38 \tab - \tab 3 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1272 \tab 267 \tab 15 \tab 41 \tab 1 \tab 3 \tab - \tab 1599 \cr
   1984 \tab 1181 \tab 233 \tab 10 \tab 47 \tab - \tab 2 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1174 \tab 245 \tab 11 \tab 36 \tab - \tab 4 \tab - \tab 1470 \cr
   1987 \tab 1430 \tab 304 \tab 24 \tab 52 \tab - \tab 9 \tab - \tab 1819 \cr
   1988 \tab 1264 \tab 174 \tab 8 \tab 32 \tab - \tab 3 \tab - \tab 1481 \cr
   1989 \tab 1334 \tab 163 \tab 14 \tab 22 \tab - \tab 4 \tab - \tab 1537 \cr
   1990 \tab 1192 \tab 148 \tab 6 \tab 24 \tab - \tab 2 \tab - \tab 1372 \cr
   1991 \tab 1310 \tab 164 \tab 5 \tab 30 \tab - \tab 8 \tab - \tab 1517 \cr
   1993 \tab 1404 \tab 155 \tab 17 \tab 29 \tab - \tab 1 \tab - \tab 1606 \cr
   1994 \tab 2887 \tab 87 \tab 7 \tab 10 \tab - \tab 1 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 65043 \tab 3073 \tab 172 \tab 503 \tab 4 \tab 51 \tab 6853 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\seealso{
Other Core: 
\code{\link{antiwar}},
\code{\link{civright}},
\code{\link{gun}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitnum}},
\code{\link{prowar}},
\code{\link{school}},
\code{\link{strike}}

Other Guns: 
\code{\link{gun}}
}
\concept{Core}
\concept{Guns}
\keyword{variable}
