% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{gunfirm}
\alias{gunfirm}
\title{How firm is r's opinion on gun control}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunfirm
}
\details{
Question How firm are you about your opinion on gun control--would you say you are very likely to change your opinion, somewhat likely to change, somewhat unlikely to change, or very unlikely to change?
}
\section{Values}{

\itemize{
\item \verb{[1]} change very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} change very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/275/vshow}{GSS Data Explorer page for \code{gunfirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab change very likely \tab somewhat likely \tab somewhat unlikely \tab change very unlikely \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 12 \tab 85 \tab 124 \tab 249 \tab 11 \tab 3 \tab 1473 \cr
   Total \tab 989 \tab 12 \tab 85 \tab 124 \tab 249 \tab 11 \tab 3 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
