% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{gunimp}
\alias{gunimp}
\title{Importance of gun control issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunimp
}
\details{
Question How important is the gun control issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most imp
\item \verb{[2]} important
\item \verb{[3]} not very imp
\item \verb{[4]} not important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1976, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/273/vshow}{GSS Data Explorer page for \code{gunimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not important \tab don't know \tab no answer \tab Total \cr
   1976 \tab - \tab 197 \tab 801 \tab 387 \tab 95 \tab 17 \tab 2 \tab 1499 \cr
   1984 \tab 989 \tab 81 \tab 252 \tab 121 \tab 20 \tab 6 \tab 4 \tab 1473 \cr
   Total \tab 989 \tab 278 \tab 1053 \tab 508 \tab 115 \tab 23 \tab 6 \tab 2972 \cr
}
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{guninfo}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
