% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_75.R
\name{gunnum}
\alias{gunnum}
\title{If threatened or shot at--how many times}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunnum
}
\details{
Question Have you ever been threatened with a gun, or shot at?
B. How many times would you guess this has happened to you?
}
\section{Values}{

\itemize{
\item \code{1} once
\item \code{2} 2-3 times
\item \code{3} 4+ times
\item \code{8} not sure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/663/vshow}{GSS Data Explorer page for \code{gunnum}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab once \tab 2-3 times \tab 4+ times \tab don't know \tab no answer \tab Total \cr
   1973 \tab 1258 \tab 144 \tab 47 \tab 38 \tab 11 \tab 6 \tab 1504 \cr
   1975 \tab 1230 \tab 142 \tab 62 \tab 35 \tab 12 \tab 9 \tab 1490 \cr
   1976 \tab 1242 \tab 134 \tab 65 \tab 52 \tab 4 \tab 2 \tab 1499 \cr
   1978 \tab 1221 \tab 187 \tab 71 \tab 40 \tab 11 \tab 2 \tab 1532 \cr
   1980 \tab 1157 \tab 173 \tab 75 \tab 50 \tab 10 \tab 3 \tab 1468 \cr
   1983 \tab 1272 \tab 180 \tab 86 \tab 52 \tab 6 \tab 3 \tab 1599 \cr
   1984 \tab 1181 \tab 170 \tab 62 \tab 52 \tab 6 \tab 2 \tab 1473 \cr
   Total \tab 8561 \tab 1130 \tab 468 \tab 319 \tab 60 \tab 27 \tab 10565 \cr
}
}

\seealso{
Other Core: 
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{polabuse}},
\code{\link{polhitok}},
\code{\link{polmurdr}}
}
\concept{Core}
\keyword{variable}
