% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{guns911}
\alias{guns911}
\title{Should gun control law be stricter after 911}
\source{
General Social Survey https://gss.norc.org
}
\description{
guns911
}
\details{
Question As a result of the 9/11 terrorist attacks, do you think that gun control laws should be stricter, making it harder for people to purchase firearms or that gun control laws should be less strict, making it easier for people to purchase firearms?
}
\section{Values}{

\itemize{
\item \code{1} should be stricter
\item \code{2} should be less strict
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3703/vshow}{GSS Data Explorer page for \code{guns911}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should be stricter \tab should be less strict \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3139 \tab 1026 \tab 226 \tab 103 \tab 16 \tab 4510 \cr
   Total \tab 3139 \tab 1026 \tab 226 \tab 103 \tab 16 \tab 4510 \cr
}
}

\seealso{
Other Gun Control: 
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{rifles50}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\keyword{variable}
