% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{gunsales}
\alias{gunsales}
\title{Should background check be required for private gun sale}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunsales
}
\details{
Question In most states a gun owner may legally sell his or her gun without proof that the buyer has passed a criminal history check. How strongly do you favor or oppose a law that required private gun sales to be subject to the same background check requirements as sales by licensed dealers? Do you…
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3700/vshow}{GSS Data Explorer page for \code{gunsales}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3139 \tab 826 \tab 236 \tab 109 \tab 91 \tab 71 \tab 26 \tab 12 \tab 4510 \cr
   Total \tab 3139 \tab 826 \tab 236 \tab 109 \tab 91 \tab 71 \tab 26 \tab 12 \tab 4510 \cr
}
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{rifles50}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\keyword{variable}
