% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{gunsdrnk}
\alias{gunsdrnk}
\title{Should carrying a firearm drinking alcohol be illegal}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunsdrnk
}
\details{
Question In all states it is illegal to drive while under the influence of alcohol. Would you favor or oppose state laws making it illegal to carry a firearm while under the influence of alcohol?
}
\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3706/vshow}{GSS Data Explorer page for \code{gunsdrnk}}.
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{rifles50}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\keyword{variable}
