% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{gunsdrug}
\alias{gunsdrug}
\title{Should penalty for illegal gun sale be tougher than drug}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunsdrug
}
\details{
Question Should penalties for illegally selling guns be tougher, less tough, or about as tough as penalties for illegally selling drugs?
}
\section{Values}{

\itemize{
\item \verb{[1]} tougher
\item \verb{[2]} less tough
\item \verb{[3]} about as tough
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3701/vshow}{GSS Data Explorer page for \code{gunsdrug}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab tougher \tab less tough \tab about as tough \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3139 \tab 734 \tab 104 \tab 490 \tab 31 \tab 12 \tab 4510 \cr
   Total \tab 3139 \tab 734 \tab 104 \tab 490 \tab 31 \tab 12 \tab 4510 \cr
}
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{othguns}},
\code{\link{rifles50}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\keyword{variable}
