% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{gvinflu2}
\alias{gvinflu2}
\title{What peo or org have 2nd most inf}
\source{
General Social Survey https://gss.norc.org
}
\description{
gvinflu2
}
\details{
Question (HAND CARD C4)
(Here is a list of people and organizations that can influence government actions.)
And which do you think has the second most influence on the actions of the American government?
}
\section{Values}{

\itemize{
\item \code{1} the media
\item \code{2} trade unions
\item \code{3} businesses, banks and industry
\item \code{4} religious organizations/authorities
\item \code{5} the military/army
\item \code{6} organized crime
\item \code{7} people who vote for the party/parties in government
\item \code{8} citizens in general
\item \code{9} civic and voluntary organizations
\item \code{10} international organizations (e.g. united nations, international monetary fund)
\item \code{11} can't choose
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6309/vshow}{GSS Data Explorer page for \code{gvinflu2}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab the media \tab international organizations (e.g. united nations, international monetary fund) \tab 100 \tab can't choose \tab trade unions \tab businesses, banks and industry \tab religious organizations/authorities \tab the military/army \tab organized crime \tab people who vote for the party/parties in government \tab citizens in general \tab civic and voluntary organizations \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 186 \tab 140 \tab 164 \tab 29 \tab 111 \tab 212 \tab 35 \tab 95 \tab 20 \tab 224 \tab 145 \tab 13 \tab 16 \tab 2867 \cr
   Total \tab 1477 \tab 186 \tab 140 \tab 164 \tab 29 \tab 111 \tab 212 \tab 35 \tab 95 \tab 20 \tab 224 \tab 145 \tab 13 \tab 16 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{polgbeco}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
