% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{gvtmelt}
\alias{gvtmelt}
\title{Should govt help blend cultures}
\source{
General Social Survey https://gss.norc.org
}
\description{
gvtmelt
}
\details{
Question Some people say that it is better for America if different racial and ethnic groups maintain their distinct cultures. Others say that it is better if groups change so that they blend into the larger society as in the idea of a melting pot. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that racial and ethnic groups should maintain their distinct cultures, and a score of 7 as meaning that groups should change so that they blend into the larger society. What score between 1 and 7 comes closest to the way you feel? (CIRCLE ONE):
B. Should it be up to the government to help racial and ethnic groups change so that they blend into the larger society, or should this be left up to the groups themselves?
}
\section{Values}{

\itemize{
\item \code{1} govt to help
\item \code{2} left up to groups
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1601/vshow}{GSS Data Explorer page for \code{gvtmelt}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{gvtapart}},
\code{\link{meltpot}}
}
\concept{Multiculturalism}
\keyword{variable}
