% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{gvtrghts}
\alias{gvtrghts}
\title{Govt should respect democratic rights}
\source{
General Social Survey https://gss.norc.org
}
\description{
gvtrghts
}
\details{
Question On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
A. That governments respect democratic rights whatever the circumstances
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4964/vshow}{GSS Data Explorer page for \code{gvtrghts}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1274 \tab 18 \tab 13 \tab 36 \tab 128 \tab 159 \tab 220 \tab 608 \tab 80 \tab 2 \tab 2538 \cr
   Total \tab 1274 \tab 18 \tab 13 \tab 36 \tab 128 \tab 159 \tab 220 \tab 608 \tab 80 \tab 2 \tab 2538 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
