% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_371.R
\name{gwpol}
\alias{gwpol}
\title{Understanding of global warming by elected officials}
\source{
General Social Survey https://gss.norc.org
}
\description{
gwpol
}
\details{
Question The first issue is global warming. Global warming means a trend toward warmer temperatures throughoutthe world, with more extreme weather in many places and changes in food production that could affect our way of life. Some people believe that the burning of gasoline and other fossil fuels causes global warming. Others say that global warming has purely natural causes. Please look at Card B2. On a scale of 1 to 5, where 1 means "Very Well" and 5 means "Not at All", how well do the following groups understand the causes of global warming?
B. Elected Officials
}
\section{Values}{

\itemize{
\item \verb{[1]} very well
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3356/vshow}{GSS Data Explorer page for \code{gwpol}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very well \tab 2 \tab 3 \tab 4 \tab not at all \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2006 \tab 3582 \tab 47 \tab 65 \tab 286 \tab 261 \tab 228 \tab 38 \tab 3 \tab - \tab 4510 \cr
   2010 \tab 1569 \tab 27 \tab 43 \tab 129 \tab 144 \tab 108 \tab 20 \tab 4 \tab - \tab 2044 \cr
   Total \tab 5151 \tab 74 \tab 108 \tab 415 \tab 405 \tab 336 \tab 58 \tab 7 \tab 0 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{advfront}},
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{architct}},
\code{\link{artseals}},
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{bettrlfe}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{biosci}},
\code{\link{bizbstgm}},
\code{\link{bizbstgw}},
\code{\link{bizbstnk}},
\code{\link{bizbsttx}},
\code{\link{bizinfgm}},
\code{\link{bizinfgw}},
\code{\link{bizinfnk}},
\code{\link{bizinftx}},
\code{\link{biznuke}},
\code{\link{boyorgrl}},
\code{\link{caremost}},
\code{\link{chemgen}},
\code{\link{cloning}},
\code{\link{cmprgmng}},
\code{\link{coldeg1}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{comorsci}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{eatgm}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{econsci}},
\code{\link{electron}},
\code{\link{engagrnk}},
\code{\link{engbrng}},
\code{\link{engbstnk}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engnrsci}},
\code{\link{engnuke}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{enhinfnk}},
\code{\link{erosion}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{extinct}},
\code{\link{farming}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{gasregs}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{gmbiz}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{goldfish}},
\code{\link{gwbiz}},
\code{\link{gwsci}},
\code{\link{h2olife}},
\code{\link{histsci}},
\code{\link{hotcore}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsmath}},
\code{\link{hsphys}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{indusgen}},
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intenvir}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intmil}},
\code{\link{intrhome}},
\code{\link{intsci}},
\code{\link{intspace}},
\code{\link{inttech}},
\code{\link{inuit}},
\code{\link{inuitway}},
\code{\link{journlsm}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lawenfrc}},
\code{\link{leadsci}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{majorcol}},
\code{\link{marrcoun}},
\code{\link{medagrgm}},
\code{\link{medagrsc}},
\code{\link{medbstgm}},
\code{\link{medbstsc}},
\code{\link{medinfgm}},
\code{\link{medinfsc}},
\code{\link{medsci}},
\code{\link{medtreat}},
\code{\link{morempg}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{newsfrmy}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{noicecap}},
\code{\link{nosun}},
\code{\link{nukeelec}},
\code{\link{nukegen}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{penguins}},
\code{\link{physcsci}},
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}},
\code{\link{polbstgm}},
\code{\link{polbstgw}},
\code{\link{polbstnk}},
\code{\link{polbstsc}},
\code{\link{polbsttx}},
\code{\link{polinfgm}},
\code{\link{polinfgw}},
\code{\link{polinfnk}},
\code{\link{polinfsc}},
\code{\link{polinftx}},
\code{\link{polnuke}},
\code{\link{radioact}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{salth2o}},
\code{\link{sciagrgw}},
\code{\link{scibnfts}},
\code{\link{scibstgw}},
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientbr}},
\code{\link{scientda}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientmo}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientre}},
\code{\link{scientsn}},
\code{\link{scientwk}},
\code{\link{scifrom}},
\code{\link{scifromy}},
\code{\link{sciimp1}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{sciinfgo}},
\code{\link{sciinfgw}},
\code{\link{sciintro}},
\code{\link{scimath}},
\code{\link{scimode}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{scispec}},
\code{\link{scistudy}},
\code{\link{scitext}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}},
\code{\link{scresrch}},
\code{\link{sealevel}},
\code{\link{seeksci}},
\code{\link{seeksciy}},
\code{\link{seesand}},
\code{\link{slsmnshp}},
\code{\link{socsci}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen}},
\code{\link{tomatoes}},
\code{\link{toofast}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}},
\code{\link{txbiz}},
\code{\link{txeco}},
\code{\link{txpol}},
\code{\link{upbreath}},
\code{\link{viruses}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}},
\code{\link{watergen}},
\code{\link{weighing}},
\code{\link{whichsci}}

Other Environment: 
\code{\link{amprogrn}},
\code{\link{anrights}},
\code{\link{anscitst}},
\code{\link{antests}},
\code{\link{artseals}},
\code{\link{bizbstgw}},
\code{\link{bizbstnk}},
\code{\link{bizinfgw}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{bizpeop}},
\code{\link{busdecid}},
\code{\link{busgrn}},
\code{\link{caremost}},
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{chemfam}},
\code{\link{chemfree}},
\code{\link{chemgen}},
\code{\link{comorsci}},
\code{\link{drivless}},
\code{\link{econgrn}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{extinct}},
\code{\link{gasregs}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grncon}},
\code{\link{grndemo}},
\code{\link{grnecon}},
\code{\link{grneffme}},
\code{\link{grnexagg}},
\code{\link{grngroup}},
\code{\link{grnintl}},
\code{\link{grnlaws}},
\code{\link{grnmoney}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grnsign}},
\code{\link{grnsol}},
\code{\link{grntaxes}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{gwbiz}},
\code{\link{gwsci}},
\code{\link{h2oless}},
\code{\link{harmsgrn}},
\code{\link{helpharm}},
\code{\link{ihlpgrn}},
\code{\link{impgrn}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{intenvir}},
\code{\link{inuitway}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{knwgw}},
\code{\link{ldcgrn}},
\code{\link{morempg}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natrecon}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{nobuygrn}},
\code{\link{noicecap}},
\code{\link{nomeat}},
\code{\link{nukeacc}},
\code{\link{nukeelec}},
\code{\link{nukefam}},
\code{\link{nukegen}},
\code{\link{othssame}},
\code{\link{penguins}},
\code{\link{peopgovt}},
\code{\link{peopgrn}},
\code{\link{polbstgw}},
\code{\link{polbstnk}},
\code{\link{polinfgw}},
\code{\link{polinfnk}},
\code{\link{polnuke}},
\code{\link{popgrwth}},
\code{\link{pubdecid}},
\code{\link{recycle}},
\code{\link{redcehme}},
\code{\link{resnatur}},
\code{\link{sciagrgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{sciworse}},
\code{\link{sealevel}},
\code{\link{spenviro}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{tempgen1}},
\code{\link{toodifme}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{usdoenuf}},
\code{\link{waterfam}},
\code{\link{watergen}}

Other Science: 
\code{\link{accntsci}},
\code{\link{advfront}},
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{artseals}},
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{bettrlfe}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{biosci}},
\code{\link{bizbstgm}},
\code{\link{bizbstgw}},
\code{\link{bizinfgm}},
\code{\link{bizinfgw}},
\code{\link{boyorgrl}},
\code{\link{caremost}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{comorsci}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{eatgm}},
\code{\link{econsci}},
\code{\link{electron}},
\code{\link{engnrsci}},
\code{\link{erosion}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{extinct}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{gmbiz}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{goldfish}},
\code{\link{gwbiz}},
\code{\link{gwsci}},
\code{\link{h2olife}},
\code{\link{harmgood}},
\code{\link{histsci}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{intspace}},
\code{\link{inuit}},
\code{\link{inuitway}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{leadsci}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{medagrgm}},
\code{\link{medagrsc}},
\code{\link{medbstgm}},
\code{\link{medbstsc}},
\code{\link{medinfgm}},
\code{\link{medinfsc}},
\code{\link{medsci}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{newsfrmy}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{noicecap}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{physcsci}},
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}},
\code{\link{polbstgm}},
\code{\link{polbstgw}},
\code{\link{polbstsc}},
\code{\link{polinfgm}},
\code{\link{polinfgw}},
\code{\link{polinfsc}},
\code{\link{radioact}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{salsci}},
\code{\link{salth2o}},
\code{\link{sciagrgw}},
\code{\link{scibnfts}},
\code{\link{scibstgw}},
\code{\link{scichng}},
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientda}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientsn}},
\code{\link{scifaith}},
\code{\link{scifrom}},
\code{\link{scifromy}},
\code{\link{scigrn}},
\code{\link{sciimp1}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{sciinfgo}},
\code{\link{sciinfgw}},
\code{\link{scimath}},
\code{\link{scimoral}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scispec}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{sciworse}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}},
\code{\link{scresrch}},
\code{\link{sealevel}},
\code{\link{seeksci}},
\code{\link{seeksciy}},
\code{\link{seesand}},
\code{\link{socsci}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{toofast}},
\code{\link{trustsci}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}},
\code{\link{upbreath}},
\code{\link{viruses}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{weighing}},
\code{\link{whichsci}}
}
\concept{Environment}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
