% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{h2oless}
\alias{h2oless}
\title{How often save water}
\source{
General Social Survey https://gss.norc.org
}
\description{
h2oless
}
\details{
Question K2. And how often do you choose to save or re-use water for environmental reasons?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4814/vshow}{GSS Data Explorer page for \code{h2oless}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   2010 \tab 614 \tab 94 \tab 308 \tab 386 \tab 631 \tab 3 \tab 8 \tab 2044 \cr
   Total \tab 614 \tab 94 \tab 308 \tab 386 \tab 631 \tab 3 \tab 8 \tab 2044 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{peopgrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}
}
\concept{ISSP Environment}
\keyword{variable}
