% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{h2olife}
\alias{h2olife}
\title{What property of water is most important}
\source{
General Social Survey https://gss.norc.org
}
\description{
h2olife
}
\details{
Question What property of water is most important for living organisms?
}
\section{Values}{

\itemize{
\item \verb{[1]} it is odorless
\item \verb{[2]} it does not conduct electricity
\item \verb{[3]} it is tasteless
\item \verb{[4]} it is liquid at most temperatures on earth
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3471/vshow}{GSS Data Explorer page for \code{h2olife}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab it is odorless \tab it does not conduct electricity \tab it is tasteless \tab it is liquid at most temperatures on earth \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 64 \tab 70 \tab 71 \tab 1012 \tab 268 \tab 20 \tab 2023 \cr
   Total \tab 518 \tab 64 \tab 70 \tab 71 \tab 1012 \tab 268 \tab 20 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
