% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{handmove}
\alias{handmove}
\title{R perform forceful hand movements}
\source{
General Social Survey https://gss.norc.org
}
\description{
handmove
}
\details{
Question Does your job regularly require you to perform repetitive or forceful hand movements or involve awkward postures?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2813/vshow}{GSS Data Explorer page for \code{handmove}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 906 \tab 871 \tab 6 \tab 13 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 833 \tab 886 \tab 3 \tab 12 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 526 \tab 634 \tab 3 \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 577 \tab 661 \tab 5 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 696 \tab 717 \tab 5 \tab 30 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 814 \tab 1142 \tab 5 \tab 12 \tab 16 \tab 3544 \cr
   Total \tab 8345 \tab 4352 \tab 4911 \tab 27 \tab 98 \tab 16 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
