% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{hapcohab}
\alias{hapcohab}
\title{Happiness of relt with partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapcohab
}
\details{
Question B. Taking things all together, would you say that your relationship with your partner is very happy, pretty happy, or not too happy?
}
\section{Values}{

\itemize{
\item \code{1} very happy
\item \code{2} pretty happy
\item \code{3} not too happy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/436/vshow}{GSS Data Explorer page for \code{hapcohab}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very happy \tab pretty happy \tab not too happy \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 1806 \tab 103 \tab 62 \tab 3 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2328 \tab 129 \tab 70 \tab 9 \tab 2 \tab - \tab - \tab 2538 \cr
   2016 \tab 2603 \tab 149 \tab 102 \tab 11 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2116 \tab 148 \tab 75 \tab 9 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3665 \tab 185 \tab 155 \tab 24 \tab 1 \tab 1 \tab 1 \tab 4032 \cr
   2022 \tab 3226 \tab 190 \tab 113 \tab 10 \tab 1 \tab 3 \tab 1 \tab 3544 \cr
   Total \tab 15744 \tab 904 \tab 577 \tab 66 \tab 5 \tab 5 \tab 2 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{fair}},
\code{\link{health}},
\code{\link{helpful}},
\code{\link{life}},
\code{\link{satcity}},
\code{\link{sathobby}},
\code{\link{trust}},
\code{\link{trusty}}
}
\concept{Core}
\keyword{variable}
