% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{hapgirls}
\alias{hapgirls}
\title{Girls' chances for a happy family life}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapgirls
}
\details{
Question Think about girls growing up today: Do you think their chances for a happy family life will be better than yours, about the same, or worse than yours?
}
\section{Values}{

\itemize{
\item \verb{[1]} better
\item \verb{[2]} about the same
\item \verb{[3]} worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1654/vshow}{GSS Data Explorer page for \code{hapgirls}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab better \tab about the same \tab worse \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 284 \tab 624 \tab 515 \tab 34 \tab 3 \tab 2904 \cr
   Total \tab 1444 \tab 284 \tab 624 \tab 515 \tab 34 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapboys}},
\code{\link{rhmewrk}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{willwed1}},
\code{\link{willwed2}}
}
\concept{Gender}
\keyword{variable}
