% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_49.R
\name{happy}
\alias{happy}
\title{General happiness}
\source{
General Social Survey https://gss.norc.org
}
\description{
happy
}
\details{
Question Taken all together, how would you say things are these days--would you say that you are very happy, pretty happy, or not too happy?
}
\section{Values}{

\itemize{
\item \verb{[1]} very happy
\item \verb{[2]} pretty happy
\item \verb{[3]} not too happy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/434/vshow}{GSS Data Explorer page for \code{happy}}.

Counts by year:\tabular{lllllllll}{
   year \tab very happy \tab pretty happy \tab not too happy \tab no answer \tab don't know \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 486 \tab 855 \tab 265 \tab 7 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 538 \tab 766 \tab 196 \tab 4 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 561 \tab 725 \tab 194 \tab 4 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 488 \tab 803 \tab 194 \tab 5 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 511 \tab 800 \tab 188 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 532 \tab 813 \tab 182 \tab 3 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 521 \tab 851 \tab 145 \tab 15 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 496 \tab 771 \tab 195 \tab 5 \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 567 \tab 1019 \tab 269 \tab 5 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 490 \tab 882 \tab 201 \tab 26 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 502 \tab 756 \tab 187 \tab 27 \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 438 \tab 918 \tab 174 \tab 4 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 468 \tab 816 \tab 165 \tab 21 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 518 \tab 1023 \tab 239 \tab 39 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 498 \tab 832 \tab 136 \tab 15 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 498 \tab 880 \tab 148 \tab 11 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 784 \tab 122 \tab 11 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 467 \tab 872 \tab 165 \tab 13 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 506 \tab 917 \tab 178 \tab 5 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 857 \tab 1756 \tab 364 \tab 14 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 877 \tab 1659 \tab 349 \tab 18 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab 891 \tab 1575 \tab 340 \tab 24 \tab 2 \tab - \tab - \tab 2832 \cr
   2000 \tab 881 \tab 1603 \tab 293 \tab 39 \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 415 \tab 784 \tab 170 \tab 2 \tab 1 \tab 1393 \tab - \tab 2765 \cr
   2004 \tab 419 \tab 738 \tab 180 \tab 1 \tab 2 \tab 1472 \tab - \tab 2812 \cr
   2006 \tab 920 \tab 1676 \tab 390 \tab - \tab 6 \tab 1518 \tab - \tab 4510 \cr
   2008 \tab 599 \tab 1100 \tab 316 \tab 6 \tab 2 \tab - \tab - \tab 2023 \cr
   2010 \tab 538 \tab 1184 \tab 317 \tab 3 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 593 \tab 1094 \tab 277 \tab 4 \tab 6 \tab - \tab - \tab 1974 \cr
   2014 \tab 786 \tab 1403 \tab 341 \tab 2 \tab 6 \tab - \tab - \tab 2538 \cr
   2016 \tab 806 \tab 1601 \tab 452 \tab 5 \tab 3 \tab - \tab - \tab 2867 \cr
   2018 \tab 701 \tab 1307 \tab 336 \tab - \tab 4 \tab - \tab - \tab 2348 \cr
   2021 \tab 783 \tab 2308 \tab 923 \tab 1 \tab 1 \tab - \tab 16 \tab 4032 \cr
   2022 \tab 779 \tab 1942 \tab 799 \tab 4 \tab 11 \tab - \tab 9 \tab 3544 \cr
   Total \tab 20385 \tab 37813 \tab 9390 \tab 343 \tab 51 \tab 4383 \tab 25 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{hapmar}}
}
\concept{Core}
\keyword{variable}
