% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{happy7}
\alias{happy7}
\title{How happy r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
happy7
}
\details{
Question If you were to consider your life in general, how happy or unhappy would you say you are, on the whole?
}
\section{Values}{

\itemize{
\item \verb{[1]} completely happy
\item \verb{[2]} very happy
\item \verb{[3]} fairly happy
\item \verb{[4]} neither happy nor unhappy
\item \verb{[5]} fairly unhappy
\item \verb{[6]} very unhappy
\item \verb{[7]} completely unhappy
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4451/vshow}{GSS Data Explorer page for \code{happy7}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab completely happy \tab very happy \tab fairly happy \tab neither happy nor unhappy \tab fairly unhappy \tab very unhappy \tab completely unhappy \tab can't choose \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 1594 \tab 141 \tab 510 \tab 391 \tab 69 \tab 32 \tab 16 \tab 1 \tab 7 \tab 4 \tab - \tab - \tab 2765 \cr
   2012 \tab 672 \tab 148 \tab 563 \tab 440 \tab 77 \tab 35 \tab 16 \tab 5 \tab 11 \tab 7 \tab - \tab - \tab 1974 \cr
   2022 \tab 1220 \tab 282 \tab 782 \tab 865 \tab 175 \tab 123 \tab 34 \tab 8 \tab 43 \tab 5 \tab 7 \tab - \tab 3544 \cr
   Total \tab 3486 \tab 571 \tab 1855 \tab 1696 \tab 321 \tab 190 \tab 66 \tab 14 \tab 61 \tab 16 \tab 7 \tab 0 \tab 8283 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{buythngs}},
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{chldcare}},
\code{\link{clean1}},
\code{\link{cohabfst}},
\code{\link{cohabit}},
\code{\link{cohabok}},
\code{\link{cooking1}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{deckids}},
\code{\link{dinner}},
\code{\link{divbest}},
\code{\link{divhubby}},
\code{\link{divifkd1}},
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokd1}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{eldcost}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{evdiv}},
\code{\link{fambudgt}},
\code{\link{famsuffr}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{fejobind}},
\code{\link{femarry}},
\code{\link{fewknokd}},
\code{\link{fewrksup}},
\code{\link{fourkids}},
\code{\link{hapifwrk}},
\code{\link{hhclean1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}},
\code{\link{kidsocst}},
\code{\link{kidsuffr}},
\code{\link{kidtrble}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{mardiv}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{mawork14}},
\code{\link{mawrkwrm}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{memarry}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{rfamlook}},
\code{\link{rhhwork}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{sexhar}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{singlpar}},
\code{\link{spborn}},
\code{\link{spevdiv}},
\code{\link{spfalook}},
\code{\link{sphhwork}},
\code{\link{splive}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{teensex1}},
\code{\link{threkids}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhm1}},
\code{\link{tiredhme}},
\code{\link{tiredwk1}},
\code{\link{tiredwrk}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{twokids}},
\code{\link{weekend}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}

Other Happiness: 
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{hapunhap}}

Other Split Ballots: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{aidsads}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blkcol}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{colaff}},
\code{\link{colop}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{father}},
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fewrite}},
\code{\link{forbdcom}},
\code{\link{forbdmar}},
\code{\link{forbdrac}},
\code{\link{form}},
\code{\link{formwt}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{healer}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{israel}},
\code{\link{japan}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{king}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{liberatr}},
\code{\link{lover}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mapa}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{meserve}},
\code{\link{mother}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{opfamily}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racimp}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{rank}()},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{russia}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{spouse}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{uswar}},
\code{\link{uswary}},
\code{\link{xmovie1}}
}
\concept{Happiness}
\concept{ISSP Family & Gender Roles}
\concept{Split Ballots}
\keyword{variable}
