% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{happy_next}
\alias{happy_next}
\title{General happiness - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
happy_next
}
\details{
Question Taken all together, how would you say things are these days—would you say that you are very happy, pretty happy, or not too happy?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} very happy
\item \verb{[2]} pretty happy
\item \verb{[3]} not too happy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8273/vshow}{GSS Data Explorer page for \code{happy_next}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab very happy \tab pretty happy \tab not too happy \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 160 \tab 622 \tab 280 \tab 2244 \tab 3 \tab 3309 \cr
   Total \tab 0 \tab 160 \tab 622 \tab 280 \tab 2244 \tab 3 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{knwsexor}},
\code{\link{perla_self}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{knwsexor}},
\code{\link{perla_self}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}
}
\concept{GSS Next}
\keyword{variable}
