% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{hapunhap}
\alias{hapunhap}
\title{Happy or unhappy with life today}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapunhap
}
\details{
Question If you were to consider your life in general these days, how happy or unhappy would you say you are, on the whole…
}
\section{Values}{

\itemize{
\item \verb{[1]} very happy
\item \verb{[2]} fairly happy
\item \verb{[3]} not very happy
\item \verb{[4]} not at all happy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998, 2002 \cr
   B/C/- \tab 2008 \cr
   C/-/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4627/vshow}{GSS Data Explorer page for \code{hapunhap}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very happy \tab fairly happy \tab not very happy \tab not at all happy \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab 158 \tab 501 \tab 737 \tab 96 \tab 12 \tab 9 \tab 4 \tab - \tab 1517 \cr
   1998 \tab 1548 \tab 467 \tab 667 \tab 113 \tab 25 \tab 6 \tab 6 \tab - \tab 2832 \cr
   2002 \tab 1616 \tab 464 \tab 575 \tab 76 \tab 14 \tab 11 \tab 9 \tab - \tab 2765 \cr
   2008 \tab 658 \tab 587 \tab 677 \tab 83 \tab 18 \tab - \tab - \tab - \tab 2023 \cr
   2018 \tab 1173 \tab 535 \tab 568 \tab 50 \tab 13 \tab 5 \tab 4 \tab - \tab 2348 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 5153 \tab 2554 \tab 3224 \tab 418 \tab 82 \tab 31 \tab 23 \tab 3544 \tab 15029 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{bestfrd}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{comyear}},
\code{\link{cousins}},
\code{\link{cowrkfrd}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{helpfrds}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{inlawnum}},
\code{\link{kidcall}},
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidpars}},
\code{\link{kidvisit}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{localgvt}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{malive}},
\code{\link{marcohab}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{neifrd}},
\code{\link{niecenep}},
\code{\link{othfrd}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{palive}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{peocntct}},
\code{\link{posslq}},
\code{\link{posslqy}},
\code{\link{relcall}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{talkedto}},
\code{\link{trustpeo}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{usefrds}},
\code{\link{wantbest}}

Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{befair}},
\code{\link{bible1}},
\code{\link{cantrust}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conbiz}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{madenkid}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relcand}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}

Other Happiness: 
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{happy7}}
}
\concept{Happiness}
\concept{ISSP Religion}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
