% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{harass5}
\alias{harass5}
\title{Harassment at work in past 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
harass5
}
\details{
Question Over the past five years, have you been harassed by your superiors or co-workers at your job, for example, have you experienced any bullying, physical or psychological abuse?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} i do not have a job/superior/co-worker
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6327/vshow}{GSS Data Explorer page for \code{harass5}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab i do not have a job/superior/co-worker \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 237 \tab 1136 \tab 96 \tab 1 \tab 7 \tab 2867 \cr
   Total \tab 1390 \tab 237 \tab 1136 \tab 96 \tab 1 \tab 7 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
