% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{harjob5}
\alias{harjob5}
\title{In last 5 yrs, r has ever been harassed on job bc r was believed to be gay/les/b}
\source{
General Social Survey https://gss.norc.org
}
\description{
harjob5
}
\details{
Question In the last five years, have you been harassed verbally or in writing on the job by a co-worker, supervisor, or boss because you were believed to be gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} not employed in last 5 years
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3925/vshow}{GSS Data Explorer page for \code{harjob5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab not employed in last 5 years \tab no answer \tab Total \cr
   2008 \tab 1971 \tab 14 \tab 4 \tab 1 \tab 33 \tab 2023 \cr
   2012 \tab 1910 \tab 13 \tab 6 \tab - \tab 45 \tab 1974 \cr
   Total \tab 3881 \tab 27 \tab 10 \tab 1 \tab 78 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
