% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{harmgood}
\alias{harmgood}
\title{Modern science does more harm than good}
\source{
General Social Survey https://gss.norc.org
}
\description{
harmgood
}
\details{
Question How much do you agree or disagree with each of these statements?
B. Overall, modern science does more harm than good.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994, 1998, 2000 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4722/vshow}{GSS Data Explorer page for \code{harmgood}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 50 \tab 200 \tab 306 \tab 686 \tab 217 \tab 38 \tab 60 \tab 1606 \cr
   1994 \tab 1606 \tab 38 \tab 165 \tab 268 \tab 628 \tab 205 \tab 38 \tab 44 \tab 2992 \cr
   1998 \tab 1548 \tab 49 \tab 127 \tab 314 \tab 487 \tab 215 \tab 58 \tab 34 \tab 2832 \cr
   2000 \tab 1541 \tab 47 \tab 185 \tab 294 \tab 457 \tab 167 \tab 37 \tab 89 \tab 2817 \cr
   2008 \tab 658 \tab 38 \tab 140 \tab 291 \tab 621 \tab 240 \tab 33 \tab 2 \tab 2023 \cr
   2010 \tab 614 \tab 41 \tab 159 \tab 314 \tab 617 \tab 227 \tab - \tab 72 \tab 2044 \cr
   Total \tab 6016 \tab 263 \tab 976 \tab 1787 \tab 3496 \tab 1271 \tab 204 \tab 301 \tab 14314 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{anrights}},
\code{\link{grnecon}},
\code{\link{harmsgrn}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scifaith}},
\code{\link{scigrn}},
\code{\link{sciworse}}
}
\concept{ISSP Environment}
\keyword{variable}
