% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{harmony}
\alias{harmony}
\title{Feel deep inner peace or harmony}
\source{
General Social Survey https://gss.norc.org
}
\description{
harmony
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
C. I feel deep inner peace or harmony.
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never/almst nevr
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2110/vshow}{GSS Data Explorer page for \code{harmony}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never/almst nevr \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 163 \tab 269 \tab 392 \tab 283 \tab 204 \tab 89 \tab 14 \tab 31 \tab 2832 \cr
   2004 \tab 1472 \tab 194 \tab 325 \tab 347 \tab 225 \tab 136 \tab 94 \tab 4 \tab 15 \tab 2812 \cr
   Total \tab 2859 \tab 357 \tab 594 \tab 739 \tab 508 \tab 340 \tab 183 \tab 18 \tab 46 \tab 5644 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{beausprt}},
\code{\link{effort}},
\code{\link{fidgety}},
\code{\link{godlove}},
\code{\link{hopeless}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{uniongod}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
