% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{harsexcl}
\alias{harsexcl}
\title{Since 18 r has ever found self the object of sexual advances from religious lead}
\source{
General Social Survey https://gss.norc.org
}
\description{
harsexcl
}
\details{
Question Now think about church, synagogue, or other religious settings. Since you turned 18, have you ever found yourself the object of sexual advances or propositions from a minister, priest, rabbi, or other clergyperson or religious leader who was not your spouse or significant other?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3936/vshow}{GSS Data Explorer page for \code{harsexcl}}.
}

\seealso{
Other Clergy/Sex: 
\code{\link{harsexjb}},
\code{\link{ownclerg}}
}
\concept{Clergy/Sex}
\keyword{variable}
