% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{harsexjb}
\alias{harsexjb}
\title{R has been the object of sexual advances/propositions from supervisor's}
\source{
General Social Survey https://gss.norc.org
}
\description{
harsexjb
}
\details{
Question Sometimes at work people find themselves the object of sexual advances or propositions from supervisors. The advances sometimes involve physical contact and sometimes just involve sexual conversations. Have you experienced such advances?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3935/vshow}{GSS Data Explorer page for \code{harsexjb}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 233 \tab 410 \tab 1349 \tab 5 \tab 26 \tab 2023 \cr
   Total \tab 233 \tab 410 \tab 1349 \tab 5 \tab 26 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{harsexcl}},
\code{\link{ownclerg}}
}
\concept{Clergy/Sex}
\keyword{variable}
