% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_326.R
\name{havchld1}
\alias{havchld1}
\title{Aged should have child}
\source{
General Social Survey https://gss.norc.org
}
\description{
havchld1
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important is it for them to...
Have a child
A. By what age should this normally occur?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2907/vshow}{GSS Data Explorer page for \code{havchld1}}.

Counts by year:\tabular{lllllllllllllllllllllllllll}{
   year \tab iap \tab 16 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 38 \tab 40 \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2065 \tab 1 \tab 13 \tab 2 \tab 29 \tab 40 \tab 27 \tab 19 \tab 27 \tab 198 \tab 39 \tab 36 \tab 53 \tab 19 \tab 126 \tab 2 \tab 11 \tab 2 \tab 1 \tab 15 \tab 1 \tab 1 \tab 5 \tab 32 \tab 1 \tab 2765 \cr
   Total \tab 2065 \tab 1 \tab 13 \tab 2 \tab 29 \tab 40 \tab 27 \tab 19 \tab 27 \tab 198 \tab 39 \tab 36 \tab 53 \tab 19 \tab 126 \tab 2 \tab 11 \tab 2 \tab 1 \tab 15 \tab 1 \tab 1 \tab 5 \tab 32 \tab 1 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{boycott}},
\code{\link{conoffcl}},
\code{\link{getmar}},
\code{\link{getmar1}},
\code{\link{havchld}},
\code{\link{protest}},
\code{\link{signpet}},
\code{\link{supfam}},
\code{\link{supfam1}}
}
\concept{Adult Transitions}
\keyword{variable}
