% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{haveinfo}
\alias{haveinfo}
\title{Enough info to get the job done}
\source{
General Social Survey https://gss.norc.org
}
\description{
haveinfo
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
D. I have enough information to get the job done
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2799/vshow}{GSS Data Explorer page for \code{haveinfo}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 1055 \tab 621 \tab 64 \tab 37 \tab 5 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 941 \tab 663 \tab 74 \tab 24 \tab 16 \tab 16 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 664 \tab 437 \tab 43 \tab 18 \tab 1 \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 732 \tab 455 \tab 45 \tab 11 \tab 1 \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 809 \tab 527 \tab 62 \tab 15 \tab 3 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 1553 \tab 1019 \tab 792 \tab 109 \tab 27 \tab 11 \tab 11 \tab 22 \tab 3544 \cr
   Total \tab 8343 \tab 5220 \tab 3495 \tab 397 \tab 132 \tab 37 \tab 103 \tab 22 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{cowrkint}},
\code{\link{fringeok}},
\code{\link{hlpequip}},
\code{\link{jobsecok}},
\code{\link{opdevel}},
\code{\link{promtefr}},
\code{\link{supcares}},
\code{\link{wkfreedm}}
}
\concept{Quality of Working Life}
\keyword{variable}
