% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{heaven}
\alias{heaven}
\title{Belief in heaven}
\source{
General Social Survey https://gss.norc.org
}
\description{
heaven
}
\details{
Question Do you believe in. . .
C. Heaven
}
\section{Values}{

\itemize{
\item \code{1} yes, definitely
\item \code{2} yes, probably
\item \code{3} no, probably not
\item \code{4} no, definitely not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4653/vshow}{GSS Data Explorer page for \code{heaven}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably not \tab no, definitely not \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 776 \tab 275 \tab 115 \tab 63 \tab 68 \tab 62 \tab 1517 \cr
   1998 \tab 1548 \tab 770 \tab 223 \tab 90 \tab 75 \tab 71 \tab 55 \tab 2832 \cr
   2008 \tab 658 \tab 853 \tab 276 \tab 117 \tab 80 \tab 33 \tab 6 \tab 2023 \cr
   2018 \tab 1173 \tab 698 \tab 224 \tab 118 \tab 101 \tab 26 \tab 8 \tab 2348 \cr
   Total \tab 3537 \tab 3097 \tab 998 \tab 440 \tab 319 \tab 198 \tab 131 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
