% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{hefinfo1}
\alias{hefinfo1}
\title{Number of household roster informant}
\source{
General Social Survey https://gss.norc.org
}
\description{
hefinfo1
}
\details{
Question Number in family of person who was household informant (i.e. who provided information on household members).
}
\section{Values}{

\itemize{
\item \verb{[1]} 1st person
\item \verb{[2]} 2nd person
\item \verb{[3]} 3rd person
\item \verb{[4]} 4th person
\item \verb{[5]} 5th person
\item \verb{[6]} 6th person
\item \verb{[7]} 7th person
\item \verb{[8]} 8th person
\item \verb{[9]} 9th person
\item \verb{[10]} 10th person
\item \verb{[11]} 11th person
\item \verb{[12]} 12th person
\item \verb{[13]} 13th person
\item \verb{[14]} 14th person
\item \verb{[22]} non-household person
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8035/vshow}{GSS Data Explorer page for \code{hefinfo1}}.

Counts by year:\tabular{lllll}{
   year \tab not available in this year \tab 1st person \tab iap \tab Total \cr
   2021 \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 3540 \tab 4 \tab 3544 \cr
   Total \tab 4032 \tab 3540 \tab 4 \tab 7576 \cr
}
}

\keyword{variable}
