% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_313.R
\name{height}
\alias{height}
\title{R is how tall}
\source{
General Social Survey https://gss.norc.org
}
\description{
height
}
\details{
Question About how tall are you without shoes? Reported in inches.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2847/vshow}{GSS Data Explorer page for \code{height}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{misswork}},
\code{\link{ntwkhard}}
}
\concept{Quality of Working Life}
\keyword{variable}
