% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_448.R
\name{helpemp}
\alias{helpemp}
\title{How successful is govt: fighting unemployment}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpemp
}
\details{
Question How successful do you think the government in America is nowadays in each of the following areas?
E. Fighting unemployment?
}
\section{Values}{

\itemize{
\item \verb{[1]} very successful
\item \verb{[2]} quite successful
\item \verb{[3]} neither successful nor unsuccessful
\item \verb{[4]} quite unsuccessful
\item \verb{[5]} very unsuccessful
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4112/vshow}{GSS Data Explorer page for \code{helpemp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very successful \tab quite successful \tab neither successful nor unsuccessful \tab quite unsuccessful \tab very unsuccessful \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 79 \tab 358 \tab 491 \tab 410 \tab 139 \tab 40 \tab 1 \tab 4510 \cr
   Total \tab 2992 \tab 79 \tab 358 \tab 491 \tab 410 \tab 139 \tab 40 \tab 1 \tab 4510 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cctv}},
\code{\link{cutdebt}},
\code{\link{emonitor}},
\code{\link{govtinfo}},
\code{\link{helpcrim}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
