% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{helpfrds}
\alias{helpfrds}
\title{Should better off ppl help less well-off friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpfrds
}
\details{
Question Please select a response to indicate how much you agree or disagree with each of the following statements.
C. People who are better off should help friends who are less well off
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4231/vshow}{GSS Data Explorer page for \code{helpfrds}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 188 \tab 470 \tab 360 \tab 91 \tab 7 \tab 24 \tab 9 \tab 2765 \cr
   2018 \tab 1175 \tab 178 \tab 503 \tab 368 \tab 95 \tab 16 \tab 10 \tab 3 \tab 2348 \cr
   Total \tab 2791 \tab 366 \tab 973 \tab 728 \tab 186 \tab 23 \tab 34 \tab 12 \tab 5113 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{demands}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
