% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{helpfulnv}
\alias{helpfulnv}
\title{People helpful or looking out for selves (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpfulnv
}
\details{
Question Would you say that most of the time people try to be helpful, or that they are mostly just looking out for themselves?
}
\section{Values}{

\itemize{
\item \verb{[1]} try to be helpful
\item \verb{[2]} looking out for themselves
\item \verb{[3]} depends (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7611/vshow}{GSS Data Explorer page for \code{helpfulnv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab try to be helpful \tab looking out for themselves \tab depends (phone mode only: volunteered) \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2719 \tab 751 \tab 550 \tab 6 \tab 2 \tab 4 \tab 4032 \cr
   2022 \tab 2962 \tab 303 \tab 275 \tab - \tab - \tab 4 \tab 3544 \cr
   Total \tab 5681 \tab 1054 \tab 825 \tab 6 \tab 2 \tab 8 \tab 7576 \cr
}
}

\keyword{variable}
