% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{helpfulv}
\alias{helpfulv}
\title{People helpful or looking out for selves (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpfulv
}
\details{
Question Would you say that most of the time people try to be helpful, or that they are mostly just looking out for themselves?
}
\section{Values}{

\itemize{
\item \code{1} try to be helpful
\item \code{2} looking out for themselves
\item \code{3} depends (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7591/vshow}{GSS Data Explorer page for \code{helpfulv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab try to be helpful \tab looking out for themselves \tab depends (phone mode only: volunteered) \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2673 \tab 449 \tab 327 \tab 582 \tab 1 \tab - \tab 4032 \cr
   2022 \tab 2974 \tab 144 \tab 153 \tab 272 \tab - \tab 1 \tab 3544 \cr
   Total \tab 5647 \tab 593 \tab 480 \tab 854 \tab 1 \tab 1 \tab 7576 \cr
}
}

\keyword{variable}
