% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{helpharm}
\alias{helpharm}
\title{Way living helps or harms}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpharm
}
\details{
Question G1. (How much do you agree or disagree with each of these statements?)
I find it hard to know whether the way I live is helpful or harmful to the environment.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4808/vshow}{GSS Data Explorer page for \code{helpharm}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 34 \tab 409 \tab 316 \tab 534 \tab 70 \tab 60 \tab 7 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 32 \tab 392 \tab 542 \tab 690 \tab 111 \tab 68 \tab 182 \tab 12 \tab 4032 \cr
   Total \tab 2617 \tab 66 \tab 801 \tab 858 \tab 1224 \tab 181 \tab 128 \tab 189 \tab 12 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{nobuygrn}},
\code{\link{peopgrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}
}
\concept{ISSP Environment}
\keyword{variable}
