% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_448.R
\name{helphlth}
\alias{helphlth}
\title{How successful is govt: providing hlth care for the sick}
\source{
General Social Survey https://gss.norc.org
}
\description{
helphlth
}
\details{
Question How successful do you think the government in America is nowadays in each of the following areas?
A. Providing health care for the sick? Would you say…
}
\section{Values}{

\itemize{
\item \code{1} very successful
\item \code{2} quite successful
\item \code{3} neither successful nor unsuccessful
\item \code{4} quite unsuccessful
\item \code{5} very unsuccessful
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4108/vshow}{GSS Data Explorer page for \code{helphlth}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very successful \tab quite successful \tab neither successful nor unsuccessful \tab quite unsuccessful \tab very unsuccessful \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 70 \tab 300 \tab 416 \tab 491 \tab 211 \tab 29 \tab 1 \tab 4510 \cr
   Total \tab 2992 \tab 70 \tab 300 \tab 416 \tab 491 \tab 211 \tab 29 \tab 1 \tab 4510 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cctv}},
\code{\link{cutdebt}},
\code{\link{emonitor}},
\code{\link{govtinfo}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helpold}},
\code{\link{helpsec}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
