% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_459.R
\name{helphwrk}
\alias{helphwrk}
\title{Helped someone with hwork during past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
helphwrk
}
\details{
Question During the past 12 months, how often have you done any of the following things for people you know personally, such as relatives, friends, neighbors or other acquaintances?
A. Helped someone outside of your household with housework or shopping
}
\section{Values}{

\itemize{
\item \verb{[1]} more than once a week
\item \verb{[2]} once a week
\item \verb{[3]} once a month
\item \verb{[4]} at least two or three times in the past year
\item \verb{[5]} once in the past year
\item \verb{[6]} not at all in the past year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4220/vshow}{GSS Data Explorer page for \code{helphwrk}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab more than once a week \tab once a week \tab once a month \tab at least two or three times in the past year \tab once in the past year \tab not at all in the past year \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 135 \tab 98 \tab 196 \tab 340 \tab 121 \tab 250 \tab 1 \tab 8 \tab 2765 \cr
   2004 \tab 1472 \tab 125 \tab 137 \tab 291 \tab 407 \tab 106 \tab 268 \tab 2 \tab 4 \tab 2812 \cr
   2012 \tab 666 \tab 58 \tab 80 \tab 175 \tab 394 \tab 120 \tab 473 \tab 1 \tab 7 \tab 1974 \cr
   2014 \tab 1264 \tab 53 \tab 97 \tab 225 \tab 377 \tab 137 \tab 379 \tab 2 \tab 4 \tab 2538 \cr
   Total \tab 5018 \tab 371 \tab 412 \tab 887 \tab 1518 \tab 484 \tab 1370 \tab 6 \tab 23 \tab 10089 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1a}},
\code{\link{borrow2a}},
\code{\link{down1a}},
\code{\link{down2a}},
\code{\link{grpchrty}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{sick1a}},
\code{\link{sick2a}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
