% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{helpjob}
\alias{helpjob}
\title{Helped somebody to find a job past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpjob
}
\details{
Question During the past 12 months, how often have you done any of the following things for people you know personally, such as relatives, friends, neighbors or other acquaintances?
D. Helped somebody to find a job
}
\section{Values}{

\itemize{
\item \verb{[1]} more than once a week
\item \verb{[2]} once a week
\item \verb{[3]} once a month
\item \verb{[4]} at least two or three times in the past year
\item \verb{[5]} once in the past year
\item \verb{[6]} not at all in the past year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4223/vshow}{GSS Data Explorer page for \code{helpjob}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab more than once a week \tab once a week \tab once a month \tab at least two or three times in the past year \tab once in the past year \tab not at all in the past year \tab no answer \tab don't know \tab Total \cr
   2002 \tab 1616 \tab 26 \tab 26 \tab 75 \tab 295 \tab 222 \tab 497 \tab 8 \tab - \tab 2765 \cr
   2004 \tab 1472 \tab 36 \tab 33 \tab 112 \tab 366 \tab 293 \tab 496 \tab 2 \tab 2 \tab 2812 \cr
   2012 \tab 666 \tab 25 \tab 28 \tab 102 \tab 342 \tab 270 \tab 533 \tab 7 \tab 1 \tab 1974 \cr
   2014 \tab 1264 \tab 33 \tab 34 \tab 92 \tab 301 \tab 272 \tab 538 \tab 4 \tab - \tab 2538 \cr
   Total \tab 5018 \tab 120 \tab 121 \tab 381 \tab 1304 \tab 1057 \tab 2064 \tab 21 \tab 3 \tab 10089 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{kidpars}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{talkedto}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
