% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_448.R
\name{helpold}
\alias{helpold}
\title{How successful is govt: decent standard of living for the old}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpold
}
\details{
Question How successful do you think the government in America is nowadays in each of the following areas?
B. Providing a decent standard of living for the old?
}
\section{Values}{

\itemize{
\item \code{1} very successful
\item \code{2} quite successful
\item \code{3} neither successful nor unsuccessful
\item \code{4} quite unsuccessful
\item \code{5} very unsuccessful
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4109/vshow}{GSS Data Explorer page for \code{helpold}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cctv}},
\code{\link{cutdebt}},
\code{\link{emonitor}},
\code{\link{govtinfo}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpsec}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
