% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{helppoor}
\alias{helppoor}
\title{Should govt improve standard of living?}
\source{
General Social Survey https://gss.norc.org
}
\description{
helppoor
}
\details{
Question I'd like to talk with you about issues some people tell us are important. Please look at CARD BC. Some people think that the government in Washington should do everything possible to improve the standard of living of all poor Americans; they are at Point 1 on this card. Other people think it is not the government's responsibility, and that each person should take care of himself; they are at Point 5.
A. Where would you place yourself on this scale, or haven't you have up your mind on this?
}
\section{Values}{

\itemize{
\item \code{1} government should help
\item \code{3} agree with both
\item \code{5} people should care for themselves
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
   1996 \tab B/C \tab partial \cr
   1998 \tab B/C \tab partial \cr
   2000 \tab B/C \tab partial \cr
   2002 \tab B/C \tab partial \cr
   2004 \tab B/C \tab partial \cr
   2006 \tab B/C \tab partial \cr
   2008 \tab B/C \tab partial \cr
   2010 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab B/C \tab partial \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/844/vshow}{GSS Data Explorer page for \code{helppoor}}.
}

\seealso{
Other Core: 
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helpsick}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{obmepax}},
\code{\link{obmewar}}

Other Federal Government: 
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helpsick}}
}
\concept{Core}
\concept{Federal Government}
\concept{Inequality, Income}
\keyword{variable}
