% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{helpusa}
\alias{helpusa}
\title{How important to help worse off ppl in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpusa
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
H. To help people in America who are worse off than yourself
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4914/vshow}{GSS Data Explorer page for \code{helpusa}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{relmeet}},
\code{\link{watchgov}}

Other Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{watchgov}}
}
\concept{Citizenship}
\concept{ISSP Citizenship}
\keyword{variable}
