% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{helpwrld}
\alias{helpwrld}
\title{How important to help worse off ppl in rest of world}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpwrld
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
I. To help people in the rest of the world who are worse off than yourself
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4915/vshow}{GSS Data Explorer page for \code{helpwrld}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   2004 \tab 1340 \tab 69 \tab 92 \tab 148 \tab 274 \tab 335 \tab 216 \tab 326 \tab 8 \tab 4 \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 81 \tab 93 \tab 124 \tab 244 \tab 262 \tab 154 \tab 271 \tab 34 \tab 1 \tab - \tab 2538 \cr
   Total \tab 2614 \tab 150 \tab 185 \tab 272 \tab 518 \tab 597 \tab 370 \tab 597 \tab 42 \tab 5 \tab 0 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{buypol}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{citviews}},
\code{\link{cntctgov}},
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{eqtreat}},
\code{\link{fixmistk}},
\code{\link{govdook}},
\code{\link{govngos}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{gvtrghts}},
\code{\link{helpusa}},
\code{\link{hlthall}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{leftrght}},
\code{\link{leftrght1}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ntcitvte}},
\code{\link{obeylaws}},
\code{\link{oppsegov}},
\code{\link{othreasn}},
\code{\link{owndecns}},
\code{\link{paytaxes}},
\code{\link{polactve}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polnews}},
\code{\link{polopts}},
\code{\link{poltalk}},
\code{\link{powerun}},
\code{\link{racmeet}},
\code{\link{refrndms}},
\code{\link{relmeet}},
\code{\link{revmeet}},
\code{\link{rghtsmin}},
\code{\link{servepeo}},
\code{\link{servpub}},
\code{\link{signdpet}},
\code{\link{solok}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usmca1}},
\code{\link{usmca2}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Citizenship: 
\code{\link{actassoc}},
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{buypol}},
\code{\link{fucitzn}},
\code{\link{helpusa}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obeylaws}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{typevisa}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{visa}},
\code{\link{voteelec}},
\code{\link{watchgov}}
}
\concept{Citizenship}
\concept{ISSP Citizenship}
\keyword{variable}
