% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{hgunlaw}
\alias{hgunlaw}
\title{Should be more restrictions on handgun}
\source{
General Social Survey https://gss.norc.org
}
\description{
hgunlaw
}
\details{
Question Please tell me whether you agree or disagree with the following statement: "There should be more legal restrictions on handguns in our society."
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3057/vshow}{GSS Data Explorer page for \code{hgunlaw}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 924 \tab 384 \tab 27 \tab 5 \tab 2812 \cr
   Total \tab 1472 \tab 924 \tab 384 \tab 27 \tab 5 \tab 2812 \cr
}
}

\seealso{
Other Guns: 
\code{\link{crimdown}},
\code{\link{crimup}},
\code{\link{hguncrim}}
}
\concept{Guns}
\keyword{variable}
