% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{hhclean1}
\alias{hhclean1}
\title{Who does household cleaning}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhclean1
}
\details{
Question In your household who does the following things...? Is it always you, usually you, about equal or both together, usually your (spouse/partner), always your (spouse/partner), or by a third person?
Does the household cleaning?
}
\section{Values}{

\itemize{
\item \verb{[1]} always me
\item \verb{[2]} usually me
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually my spouse/partner
\item \verb{[5]} always my spouse/partner
\item \verb{[6]} is done by a third person
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4473/vshow}{GSS Data Explorer page for \code{hhclean1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab always me \tab usually me \tab about equal or both together \tab usually my spouse/partner \tab always my spouse/partner \tab is done by a third person \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2012 \tab 1274 \tab 165 \tab 124 \tab 194 \tab 116 \tab 55 \tab 43 \tab 3 \tab - \tab - \tab 1974 \cr
   2022 \tab 3024 \tab 62 \tab 94 \tab 189 \tab 101 \tab 31 \tab 33 \tab 1 \tab 7 \tab 2 \tab 3544 \cr
   Total \tab 4298 \tab 227 \tab 218 \tab 383 \tab 217 \tab 86 \tab 76 \tab 4 \tab 7 \tab 2 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
