% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{hhtype}
\alias{hhtype}
\title{Household type}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype
}
\details{
Question Household Type.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1adlt,0kids
\item \verb{[2]} 1adlt,1+kids
\item \verb{[3]} 2adlts,mar,0kids
\item \verb{[4]} 2adlts,mar,1+kids
\item \verb{[5]} 2adlts,dkmar,0kids
\item \verb{[6]} 2adlts,dkmar,1+kids
\item \verb{[7]} 2as,ntmar,ntrel,opsex,0ks
\item \verb{[8]} 2as,ntmar,ntrel,opsex,1+ks
\item \verb{[9]} 2as,ntmar,ntrel,smsex,0ks
\item \verb{[10]} 2as,ntmar,ntrel,smsex,1+ks
\item \verb{[11]} 2adlts,ntmar,rel,0kids
\item \verb{[12]} 2adlts,ntmar,rel,1+kids
\item \verb{[13]} 2adlts,ntmar,dkrel,0kids
\item \verb{[14]} 2adlts,ntmar,dkrel,1+kids
\item \verb{[15]} 2as,ntmar,ntrel,dksex,0ks
\item \verb{[16]} 2as,ntmar,ntrel,dksex,1+ks
\item \verb{[20]} 3adlt,0mar,sp,0kids
\item \verb{[21]} 3adlts,0mar,sp,rel,0kids
\item \verb{[22]} 3adlts,0mar,rel,sp,0kids
\item \verb{[23]} 3adlts,0mar,rel,0kids
\item \verb{[24]} 3adlts,0mar,sp,ntrel,0kids
\item \verb{[25]} 3adlts,0mar,ntrel,sp,0kids
\item \verb{[26]} 3adlts,0mar,rel,0kids
\item \verb{[27]} 3adlts,0mar,ntrel,rel,0kids
\item \verb{[28]} 3adlts,0mar,ntrel,0kids
\item \verb{[30]} 3adlts,1mar,sp,0kids
\item \verb{[31]} 3adlts,1mar,sp,rel,0kids
\item \verb{[32]} 3adlts,1mar,rel,sp,0kids
\item \verb{[33]} 3adlts,1mar,rel,rel,0kids
\item \verb{[34]} 3adlts,1mar,sp,ntrel,0kids
\item \verb{[35]} 3adlts,1mar,ntrel,sp,0kids
\item \verb{[36]} 3adlts,1mar,rel,ntrel,0kids
\item \verb{[37]} 3adlts,1mar,ntrel,rel,0kids
\item \verb{[38]} 3adlts,1mar,ntrel,0kids
\item \verb{[40]} 3adlts,2mar,sp,0kids
\item \verb{[41]} 3adlts,2mar,sp,rel,0kids
\item \verb{[42]} 3adlts,2mar,rel,sp,0kids
\item \verb{[43]} 3adlts,2mar,rel,0kids
\item \verb{[44]} 3adlts,2mar,sp,ntrel,0kids
\item \verb{[45]} 3adlts,2mar,rel,ntrel,0kids
\item \verb{[46]} 3adlts,2mar,rel,ntrel,0kids
\item \verb{[47]} 3adlts,2mar,ntrel,rel,0kids
\item \verb{[48]} 3adlts,2mar,ntrel,0kids
\item \verb{[50]} 3adlts,3mar,sp,0kids
\item \verb{[51]} 3adlts,3mar,sp,rel,0kids
\item \verb{[52]} 3adlts,3mar,sp,rel,0kids
\item \verb{[53]} 3adlts,3mar,rel,0kids
\item \verb{[54]} 3adlts,3mar,sp,ntrel,0kids
\item \verb{[55]} 3adlts,3mar,ntrel,sp,0kids
\item \verb{[56]} 3adlts,3mar,rel,ntrel,0kids
\item \verb{[57]} 3adlts,3mar,ntrel,rel,0kids
\item \verb{[58]} 3adlts,3mar,ntrel,0kids
\item \verb{[100]} 4+adlts,0mar,0kids
\item \verb{[101]} 4+adlts,1mar,0kids
\item \verb{[102]} 4+adlts,2mar,0kids
\item \verb{[103]} 4+adlts,3mar,0kids
\item \verb{[104]} 4+adlts,4+mar,0kids
\item \verb{[120]} 3adlt,0mar,sp,1+kids
\item \verb{[121]} 3adlts,0mar,sp,rel,1+kids
\item \verb{[122]} 3adlts,0mar,rel,sp,1+kids
\item \verb{[123]} 3adlts,0mar,rel,1+kids
\item \verb{[124]} 3adlts,0mar,sp,ntrel,1+kids
\item \verb{[125]} 3adlts,0mar,ntrel,sp,1+kids
\item \verb{[126]} 3adlts,0mar,rel,1+kids
\item \verb{[127]} 3adlts,0mar,ntrel,rel,1+kids
\item \verb{[128]} 3adlts,0mar,ntrel,1+kids
\item \verb{[130]} 3adlts,1mar,sp,1+kids
\item \verb{[131]} 3adlts,1mar,sp,rel,1+kids
\item \verb{[132]} 3adlts,1mar,rel,sp,1+kids
\item \verb{[133]} 3adlts,1mar,rel,rel,1+kids
\item \verb{[134]} 3adlts,1mar,sp,ntrel,1+kids
\item \verb{[135]} 3adlts,1mar,ntrel,sp,1+kids
\item \verb{[136]} 3adlts,1mar,rel,ntrel,1+kids
\item \verb{[137]} 3adlts,1mar,ntrel,rel,1+kids
\item \verb{[138]} 3adlts,1mar,ntrel,1+kids
\item \verb{[140]} 3adlts,2mar,sp,1+kids
\item \verb{[141]} 3adlts,2mar,sp,rel,1+kids
\item \verb{[142]} 3adlts,2mar,rel,sp,1+kids
\item \verb{[143]} 3adlts,2mar,rel,1+kids
\item \verb{[144]} 3adlts,2mar,sp,ntrel,1+kids
\item \verb{[145]} 3adlts,2mar,rel,ntrel,1+kids
\item \verb{[146]} 3adlts,2mar,rel,ntrel,1+kids
\item \verb{[147]} 3adlts,2mar,ntrel,rel,1+kids
\item \verb{[148]} 3adlts,2mar,ntrel,1+kids
\item \verb{[150]} 3adlts,3mar,sp,1+kids
\item \verb{[151]} 3adlts,3mar,sp,rel,1+kids
\item \verb{[152]} 3adlts,3mar,sp,rel,1+kids
\item \verb{[153]} 3adlts,3mar,rel,1+kids
\item \verb{[154]} 3adlts,3mar,sp,ntrel,1+kids
\item \verb{[155]} 3adlts,3mar,ntrel,sp,1+kids
\item \verb{[156]} 3adlts,3mar,rel,ntrel,1+kids
\item \verb{[157]} 3adlts,3mar,ntrel,rel,1+kids
\item \verb{[158]} 3adlts,3mar,ntrel,1+kids
\item \verb{[200]} 4+adlts,0mar,1+kids
\item \verb{[201]} 4+adlts,1mar,1+kids
\item \verb{[202]} 4+adlts,2mar,1+kids
\item \verb{[203]} 4+adlts,3mar,1+kids
\item \verb{[204]} 4+adlts,4+mar,1+kids
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{famgen}},
\code{\link{famgen_7522}},
\code{\link{famgen_exp}},
\code{\link{famgen_exp_22}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{hhtype1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relhh1}},
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh13}},
\code{\link{relhh14}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp11}},
\code{\link{relsp12}},
\code{\link{relsp13}},
\code{\link{relsp14}},
\code{\link{relsp2}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}},
\code{\link{respnum}},
\code{\link{rplace}},
\code{\link{rplace_7522}},
\code{\link{rvisitor}},
\code{\link{visitors}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}}

Other Household: 
\code{\link{adults}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{dinner}},
\code{\link{earnrs}},
\code{\link{fairhwrk}},
\code{\link{fambudgt}},
\code{\link{famfinan}},
\code{\link{famgen}},
\code{\link{famgen_7522}},
\code{\link{famgen_exp}},
\code{\link{famgen_exp_22}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{fewrksup}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{hefinfo}},
\code{\link{hhclean1}},
\code{\link{hhrace}},
\code{\link{hhtype1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homeless}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{hubbywk1}},
\code{\link{hwduties}},
\code{\link{incdef}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{mrmom}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{ownhh}},
\code{\link{ownhh1}},
\code{\link{povline}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relhh1}},
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh14}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp11}},
\code{\link{relsp12}},
\code{\link{relsp13}},
\code{\link{relsp14}},
\code{\link{relsp2}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{respnum}},
\code{\link{rhhwork}},
\code{\link{rhmewrk}},
\code{\link{rplace}},
\code{\link{rplace_7522}},
\code{\link{rvisitor}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{sphhwork}},
\code{\link{sphmewrk}},
\code{\link{strsshme}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{timehome}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{visitors}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{womenben}},
\code{\link{womenhrt}}
}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
