% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{hhtype1}
\alias{hhtype1}
\title{Household type (condensed)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype1
}
\details{
Question Household type (condensed)
}
\section{Values}{

\itemize{
\item \verb{[1]} married couple, no children
\item \verb{[2]} single parent
\item \verb{[3]} other family, no children
\item \verb{[4]} single adult
\item \verb{[5]} cohabitating couple, no children
\item \verb{[6]} non-family, no children
\item \verb{[8]} unsure, no children
\item \verb{[11]} married couple with children
\item \verb{[13]} other family with children
\item \verb{[15]} cohabitating couple with children
\item \verb{[16]} non-family with children
\item \verb{[18]} unsure with children
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5434/vshow}{GSS Data Explorer page for \code{hhtype1}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab married couple, no children \tab single parent \tab other family, no children \tab single adult \tab cohabitating couple, no children \tab non-family, no children \tab unsure, no children \tab married couple with children \tab other family with children \tab cohabitating couple with children \tab non-family with children \tab unsure with children \tab no answer \tab Total \cr
   1975 \tab - \tab 471 \tab 60 \tab 56 \tab 200 \tab 14 \tab 16 \tab 8 \tab 610 \tab 28 \tab 10 \tab 3 \tab 4 \tab 10 \tab 1490 \cr
   1976 \tab 755 \tab 227 \tab 45 \tab 23 \tab 138 \tab 7 \tab 10 \tab 5 \tab 274 \tab 12 \tab - \tab 1 \tab 2 \tab - \tab 1499 \cr
   1977 \tab - \tab 480 \tab 85 \tab 46 \tab 263 \tab 12 \tab 23 \tab 12 \tab 564 \tab 33 \tab 2 \tab 1 \tab 6 \tab 3 \tab 1530 \cr
   1978 \tab - \tab 437 \tab 79 \tab 49 \tab 297 \tab 13 \tab 26 \tab 4 \tab 573 \tab 38 \tab 5 \tab 1 \tab 5 \tab 5 \tab 1532 \cr
   1980 \tab - \tab 448 \tab 85 \tab 53 \tab 287 \tab 18 \tab 26 \tab 13 \tab 481 \tab 33 \tab - \tab 4 \tab 13 \tab 7 \tab 1468 \cr
   1982 \tab - \tab 528 \tab 134 \tab 74 \tab 410 \tab 30 \tab 29 \tab 19 \tab 540 \tab 62 \tab 9 \tab 2 \tab 18 \tab 5 \tab 1860 \cr
   1983 \tab - \tab 482 \tab 95 \tab 56 \tab 309 \tab 20 \tab 21 \tab 10 \tab 540 \tab 38 \tab 5 \tab 4 \tab 16 \tab 3 \tab 1599 \cr
   1984 \tab - \tab 433 \tab 91 \tab 46 \tab 330 \tab 18 \tab 39 \tab 17 \tab 446 \tab 29 \tab 13 \tab 2 \tab 8 \tab 1 \tab 1473 \cr
   1985 \tab - \tab 462 \tab 91 \tab 55 \tab 342 \tab 20 \tab 49 \tab 11 \tab 459 \tab 22 \tab 8 \tab 2 \tab 8 \tab 5 \tab 1534 \cr
   1986 \tab - \tab 421 \tab 71 \tab 61 \tab 313 \tab 18 \tab 36 \tab 28 \tab 459 \tab 33 \tab 10 \tab 2 \tab 13 \tab 5 \tab 1470 \cr
   1987 \tab - \tab 465 \tab 109 \tab 92 \tab 420 \tab 30 \tab 52 \tab 32 \tab 502 \tab 56 \tab 12 \tab 4 \tab 30 \tab 15 \tab 1819 \cr
   1988 \tab - \tab 425 \tab 115 \tab 52 \tab 326 \tab 31 \tab 49 \tab 13 \tab 407 \tab 38 \tab 11 \tab 1 \tab 8 \tab 5 \tab 1481 \cr
   1989 \tab - \tab 440 \tab 93 \tab 55 \tab 327 \tab 32 \tab 45 \tab 26 \tab 457 \tab 29 \tab 14 \tab 1 \tab 6 \tab 12 \tab 1537 \cr
   1990 \tab - \tab 407 \tab 79 \tab 67 \tab 329 \tab 36 \tab 28 \tab 17 \tab 360 \tab 30 \tab 11 \tab 2 \tab 1 \tab 5 \tab 1372 \cr
   1991 \tab - \tab 434 \tab 103 \tab 53 \tab 375 \tab 35 \tab 35 \tab 18 \tab 404 \tab 31 \tab 12 \tab 3 \tab 8 \tab 6 \tab 1517 \cr
   1993 \tab - \tab 469 \tab 97 \tab 70 \tab 376 \tab 43 \tab 33 \tab 14 \tab 439 \tab 27 \tab 23 \tab 5 \tab 8 \tab 2 \tab 1606 \cr
   1994 \tab - \tab 845 \tab 211 \tab 119 \tab 759 \tab 74 \tab 58 \tab 25 \tab 777 \tab 54 \tab 43 \tab 8 \tab 13 \tab 6 \tab 2992 \cr
   1996 \tab - \tab 748 \tab 238 \tab 101 \tab 725 \tab 74 \tab 75 \tab 81 \tab 687 \tab 64 \tab 50 \tab 9 \tab 26 \tab 26 \tab 2904 \cr
   1998 \tab - \tab 754 \tab 181 \tab 107 \tab 771 \tab 64 \tab 74 \tab 81 \tab 630 \tab 61 \tab 46 \tab 7 \tab 40 \tab 16 \tab 2832 \cr
   2000 \tab - \tab 652 \tab 219 \tab 139 \tab 740 \tab 100 \tab 80 \tab 33 \tab 677 \tab 71 \tab 57 \tab 12 \tab 21 \tab 16 \tab 2817 \cr
   2002 \tab - \tab 784 \tab 175 \tab 139 \tab 870 \tab 85 \tab 65 \tab 31 \tab 471 \tab 60 \tab 61 \tab 9 \tab 7 \tab 8 \tab 2765 \cr
   2004 \tab - \tab 939 \tab 169 \tab 121 \tab 706 \tab 89 \tab 49 \tab 32 \tab 590 \tab 38 \tab 46 \tab 4 \tab 20 \tab 9 \tab 2812 \cr
   2006 \tab - \tab 1309 \tab 270 \tab 198 \tab 1220 \tab 139 \tab 101 \tab 45 \tab 970 \tab 98 \tab 88 \tab 9 \tab 43 \tab 20 \tab 4510 \cr
   2008 \tab - \tab 577 \tab 148 \tab 93 \tab 520 \tab 58 \tab 44 \tab 25 \tab 424 \tab 60 \tab 43 \tab 6 \tab 16 \tab 9 \tab 2023 \cr
   2010 \tab - \tab 582 \tab 127 \tab 115 \tab 601 \tab 79 \tab 21 \tab 39 \tab 342 \tab 58 \tab 40 \tab 8 \tab 18 \tab 14 \tab 2044 \cr
   2012 \tab - \tab 594 \tab 105 \tab 142 \tab 529 \tab 80 \tab 10 \tab 24 \tab 369 \tab 45 \tab 49 \tab 1 \tab 18 \tab 8 \tab 1974 \cr
   2014 \tab - \tab 839 \tab 133 \tab 199 \tab 695 \tab 98 \tab 17 \tab 49 \tab 379 \tab 46 \tab 47 \tab 2 \tab 21 \tab 13 \tab 2538 \cr
   2016 \tab - \tab 826 \tab 144 \tab 249 \tab 850 \tab 130 \tab 17 \tab 46 \tab 440 \tab 61 \tab 66 \tab 1 \tab 24 \tab 13 \tab 2867 \cr
   2018 \tab - \tab 705 \tab 106 \tab 135 \tab 669 \tab 120 \tab 55 \tab 61 \tab 343 \tab 65 \tab 42 \tab 7 \tab 23 \tab 17 \tab 2348 \cr
   2021 \tab - \tab 1175 \tab 151 \tab 12 \tab 1021 \tab 163 \tab 29 \tab 42 \tab 610 \tab 10 \tab 70 \tab 3 \tab 18 \tab 728 \tab 4032 \cr
   2022 \tab 1764 \tab 451 \tab 91 \tab 80 \tab 594 \tab 82 \tab 28 \tab 11 \tab 300 \tab 47 \tab 43 \tab 12 \tab 28 \tab 13 \tab 3544 \cr
   2024 \tab 1630 \tab 400 \tab 96 \tab 82 \tab 577 \tab 78 \tab 19 \tab 11 \tab 290 \tab 52 \tab 37 \tab 8 \tab 20 \tab 9 \tab 3309 \cr
   Total \tab 4149 \tab 19209 \tab 3996 \tab 2939 \tab 16889 \tab 1890 \tab 1259 \tab 883 \tab 15814 \tab 1429 \tab 973 \tab 144 \tab 510 \tab 1014 \tab 71098 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{famgen}},
\code{\link{famgen_7522}},
\code{\link{famgen_exp}},
\code{\link{famgen_exp_22}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{hhtype}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relhh1}},
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh13}},
\code{\link{relhh14}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp11}},
\code{\link{relsp12}},
\code{\link{relsp13}},
\code{\link{relsp14}},
\code{\link{relsp2}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}},
\code{\link{respnum}},
\code{\link{rplace}},
\code{\link{rplace_7522}},
\code{\link{rvisitor}},
\code{\link{visitors}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}}

Other Household: 
\code{\link{adults}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{dinner}},
\code{\link{earnrs}},
\code{\link{fairhwrk}},
\code{\link{fambudgt}},
\code{\link{famfinan}},
\code{\link{famgen}},
\code{\link{famgen_7522}},
\code{\link{famgen_exp}},
\code{\link{famgen_exp_22}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{fewrksup}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{hefinfo}},
\code{\link{hhclean1}},
\code{\link{hhrace}},
\code{\link{hhtype}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homeless}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{hubbywk1}},
\code{\link{hwduties}},
\code{\link{incdef}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{mrmom}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{ownhh}},
\code{\link{ownhh1}},
\code{\link{povline}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relhh1}},
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh14}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp11}},
\code{\link{relsp12}},
\code{\link{relsp13}},
\code{\link{relsp14}},
\code{\link{relsp2}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{respnum}},
\code{\link{rhhwork}},
\code{\link{rhmewrk}},
\code{\link{rplace}},
\code{\link{rplace_7522}},
\code{\link{rvisitor}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{sphhwork}},
\code{\link{sphmewrk}},
\code{\link{strsshme}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{timehome}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{visitors}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{womenben}},
\code{\link{womenhrt}}
}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
