% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{hhtype1_exp}
\alias{hhtype1_exp}
\title{Household type (condensed) (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype1_exp
}
\details{
Question Recoded from experimental Household Composition variables ADULTS_EXP to RNRLTU18
}
\section{Values}{

\itemize{
\item \verb{[1]} married couple, no children
\item \verb{[2]} single parent
\item \verb{[3]} other family, no children
\item \verb{[4]} single adult
\item \verb{[5]} cohabitating couple, no children
\item \verb{[6]} non-family, no children
\item \verb{[8]} unsure, no children
\item \verb{[11]} married couple with children
\item \verb{[13]} other family with children
\item \verb{[15]} cohabitating couple with children
\item \verb{[16]} non-family with children
\item \verb{[18]} unsure with children
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8036/vshow}{GSS Data Explorer page for \code{hhtype1_exp}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab not available in this year \tab married couple, no children \tab single parent \tab other family, no children \tab single adult \tab cohabitating couple, no children \tab non-family, no children \tab unsure, no children \tab married couple with children \tab other family with children \tab cohabitating couple with children \tab non-family with children \tab unsure with children \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 382 \tab 132 \tab 128 \tab 563 \tab 92 \tab 22 \tab 28 \tab 269 \tab 55 \tab 60 \tab 2 \tab 9 \tab 1794 \tab 6 \tab 2 \tab 3544 \cr
   2024 \tab - \tab 380 \tab 95 \tab 108 \tab 525 \tab 73 \tab 27 \tab 32 \tab 222 \tab 65 \tab 55 \tab 2 \tab 20 \tab 1684 \tab 3 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 762 \tab 227 \tab 236 \tab 1088 \tab 165 \tab 49 \tab 60 \tab 491 \tab 120 \tab 115 \tab 4 \tab 29 \tab 3478 \tab 9 \tab 20 \tab 6853 \cr
}
}

\keyword{variable}
