% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_690.R
\name{hhtype1_next}
\alias{hhtype1_next}
\title{Household type (condensed) - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype1_next
}
\details{
Question Given what you said, which of these matches your household:
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} married couple, no children
\item \verb{[2]} single parent
\item \verb{[3]} other family, no children
\item \verb{[4]} single adult
\item \verb{[5]} cohabitating couple, no children
\item \verb{[6]} non-family, no children
\item \verb{[7]} married couple, children
\item \verb{[8]} cohabiting couple, children
\item \verb{[9]} other family relationship, children
\item \verb{[10]} non-family relationship, children
\item \verb{[11]} none of these
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8274/vshow}{GSS Data Explorer page for \code{hhtype1_next}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab not available in this year \tab married couple, no children \tab single parent \tab other family, no children \tab single adult \tab cohabitating couple, no children \tab non-family, no children \tab married couple, children \tab cohabiting couple, children \tab other family relationship, children \tab non-family relationship, children \tab none of these \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 260 \tab 87 \tab 28 \tab 260 \tab 57 \tab 13 \tab 212 \tab 40 \tab 20 \tab 3 \tab 53 \tab 2268 \tab 8 \tab 3309 \cr
   Total \tab 0 \tab 260 \tab 87 \tab 28 \tab 260 \tab 57 \tab 13 \tab 212 \tab 40 \tab 20 \tab 3 \tab 53 \tab 2268 \tab 8 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{famgen_next}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}}

Other GSS Next: 
\code{\link{famgen_next}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}}
}
\concept{GSS Next}
\keyword{variable}
