% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{hhwkdis}
\alias{hhwkdis}
\title{How often r and spouse disagree sharing of hh wrk}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhwkdis
}
\details{
Question How often do you and your spouse/partner disagree about the sharing of household work?
}
\section{Values}{

\itemize{
\item \code{1} several times a week
\item \code{2} several times a month
\item \code{3} several times a year
\item \code{4} less often or rarely
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4438/vshow}{GSS Data Explorer page for \code{hhwkdis}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{deckids}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{spborn}},
\code{\link{sphhwork}}

Other Household: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{sphhwork}}
}
\concept{Household}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
