% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{hhwkfair}
\alias{hhwkfair}
\title{Sharing of hh work between r and spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhwkfair
}
\details{
Question Which of the following best applies to the sharing of household work between you and your spouse/partner?
}
\section{Values}{

\itemize{
\item \verb{[1]} i do much more than my fair share of hh work
\item \verb{[2]} i do a bit more than my fair share of hh work
\item \verb{[3]} i do roughly my fair share of hh work
\item \verb{[4]} i do a bit less than my fair share of hh work
\item \verb{[5]} i do much less than my fair share of hh work
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4437/vshow}{GSS Data Explorer page for \code{hhwkfair}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab i do much more than my fair share of hh work \tab i do a bit more than my fair share of hh work \tab i do roughly my fair share of hh work \tab i do a bit less than my fair share of hh work \tab i do much less than my fair share of hh work \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2058 \tab 139 \tab 136 \tab 283 \tab 87 \tab 35 \tab 3 \tab 24 \tab 2765 \cr
   2012 \tab 1274 \tab 149 \tab 129 \tab 253 \tab 105 \tab 56 \tab 4 \tab 4 \tab 1974 \cr
   Total \tab 3332 \tab 288 \tab 265 \tab 536 \tab 192 \tab 91 \tab 7 \tab 28 \tab 4739 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{deckids}},
\code{\link{hhwkdis}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{spborn}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
