% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{hiinc}
\alias{hiinc}
\title{Importance of high pr income in a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
hiinc
}
\details{
Question On the following list there are various aspects of jobs. Please circle one number to show how important you personally consider it is in a job:
B. High income?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} neither important nor unimportant
\item \verb{[4]} not important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4508/vshow}{GSS Data Explorer page for \code{hiinc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab neither important nor unimportant \tab not important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 347 \tab 822 \tab 193 \tab 38 \tab 4 \tab 12 \tab 37 \tab 1537 \cr
   1998 \tab 1609 \tab 271 \tab 656 \tab 181 \tab 52 \tab 5 \tab 17 \tab 41 \tab 2832 \cr
   2006 \tab 2992 \tab 454 \tab 747 \tab 214 \tab 85 \tab 12 \tab 5 \tab 1 \tab 4510 \cr
   2016 \tab 1390 \tab 426 \tab 747 \tab 222 \tab 68 \tab 8 \tab 3 \tab 3 \tab 2867 \cr
   Total \tab 6075 \tab 1498 \tab 2972 \tab 810 \tab 243 \tab 29 \tab 37 \tab 82 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{bosswrks}},
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}},
\code{\link{strngun}}
}
\concept{Work Orientation}
\keyword{variable}
