% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{hinumoky}
\alias{hinumoky}
\title{Number hispanics in military-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
hinumoky
}
\details{
Question A. At the present time, about 4 percent of the army are Hispanics. All things considered, do you think there are too many Hispanics in the army, about the right number, or should there be more Hispanics in the army?
RECORD UNDER "A" BELOW.
}
\section{Values}{

\itemize{
\item \verb{[1]} too many
\item \verb{[2]} about right number
\item \verb{[3]} should be more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/786/vshow}{GSS Data Explorer page for \code{hinumoky}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too many \tab about right number \tab should be more \tab don't know \tab no answer \tab Total \cr
   1982 \tab 923 \tab 55 \tab 445 \tab 277 \tab 150 \tab 10 \tab 1860 \cr
   Total \tab 923 \tab 55 \tab 445 \tab 277 \tab 150 \tab 10 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
