% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{hispanic}
\alias{hispanic}
\title{Hispanic specified}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispanic
}
\details{
Question IF R IS FEMALE, READ LATINA; IF MALE, READ LATINO.
Are you Spanish, Hispanic, or Latino/Latina?
Yes . . . . . (ASK A)
No   . . . . . . (GO TO NEXT QUESTION)
IF YES:                  Which group are you from?
Mexican, Mexican American, Chicano/Chicana
Puerto Rican
Cuban
Other
}
\section{Values}{

\itemize{
\item \verb{[1]} not hispanic
\item \verb{[2]} mexican, mexican american, chicano/a
\item \verb{[3]} puerto rican
\item \verb{[4]} cuban
\item \verb{[5]} another hispanic, latino, or spanish origin
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5276/vshow}{GSS Data Explorer page for \code{hispanic}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab not hispanic \tab mexican, mexican american, chicano/a \tab puerto rican \tab cuban \tab another hispanic, latino, or spanish origin \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   2000 \tab - \tab 2604 \tab 110 \tab 44 \tab 6 \tab 52 \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2561 \tab 105 \tab 39 \tab 7 \tab 51 \tab 2 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2563 \tab 156 \tab 33 \tab 7 \tab 47 \tab 3 \tab 3 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 3922 \tab 370 \tab 69 \tab 23 \tab 124 \tab 2 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1782 \tab 148 \tab 24 \tab 15 \tab 50 \tab 1 \tab 3 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1803 \tab 147 \tab 27 \tab 10 \tab 51 \tab 3 \tab 3 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1701 \tab 168 \tab 22 \tab 9 \tab 70 \tab 3 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2139 \tab 241 \tab 42 \tab 20 \tab 90 \tab 3 \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2496 \tab 216 \tab 46 \tab 18 \tab 89 \tab - \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 1984 \tab 224 \tab 38 \tab 15 \tab 72 \tab 8 \tab 7 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 3544 \tab 245 \tab 51 \tab 21 \tab 137 \tab 1 \tab 2 \tab 21 \tab 10 \tab 4032 \cr
   2022 \tab - \tab 2930 \tab 311 \tab 68 \tab 33 \tab 165 \tab 17 \tab 5 \tab - \tab 15 \tab 3544 \cr
   Total \tab 0 \tab 30029 \tab 2441 \tab 503 \tab 184 \tab 998 \tab 44 \tab 29 \tab 21 \tab 25 \tab 34274 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethadapt}},
\code{\link{ethclose}},
\code{\link{ethdiff}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethnic}},
\code{\link{ethnofit}},
\code{\link{ethnum}},
\code{\link{ethorgs}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{fucitzn}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hispanic_0022}},
\code{\link{hspgovt}},
\code{\link{meltpot}},
\code{\link{natborn}},
\code{\link{owneth}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{schleth}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{teacheth}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{whoteach}},
\code{\link{whtgovt}}

Other Ethnicity: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethadapt}},
\code{\link{ethclose}},
\code{\link{ethdiff}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethnic}},
\code{\link{ethnofit}},
\code{\link{ethnum}},
\code{\link{ethorgs}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{fucitzn}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hispanic_0022}},
\code{\link{hspgovt}},
\code{\link{meltpot}},
\code{\link{natborn}},
\code{\link{owneth}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{schleth}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{teacheth}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{whoteach}},
\code{\link{whtgovt}}

Other Race: 
\code{\link{admirblk}},
\code{\link{allowmar}},
\code{\link{asngovt}},
\code{\link{blkgains}},
\code{\link{blkgovt}},
\code{\link{blkinflu}},
\code{\link{blksimp}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{bwfriend}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{color}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{dsndntpay}},
\code{\link{ethchng}},
\code{\link{ethhist}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdmar}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{hispanic_0022}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{hseth}},
\code{\link{hspgovt}},
\code{\link{hswht}},
\code{\link{leastcom}},
\code{\link{meltpot}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{numbwfrd}},
\code{\link{oprace}},
\code{\link{owneth}},
\code{\link{perla}},
\code{\link{perla_self}},
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs16}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs7}},
\code{\link{raceacs9}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmeet}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpromo}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{ractrust}},
\code{\link{racwork}},
\code{\link{ratetone}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{trtmyrac}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wrkwayup}}
}
\concept{Ethnicity}
\concept{Race}
\keyword{variable}
