% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_260.R
\name{hispcls}
\alias{hispcls}
\title{Does r feel close to hispanics}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispcls
}
\details{
Question Now, I'm going to ask some questions about these people you personally know.
E. Are any of these Hispanics people that you feel close to?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2354/vshow}{GSS Data Explorer page for \code{hispcls}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1878 \tab 401 \tab 408 \tab 10 \tab 120 \tab 2817 \cr
   Total \tab 1878 \tab 401 \tab 408 \tab 10 \tab 120 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispwrk}},
\code{\link{mostcom}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
