% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_260.R
\name{hispwrk}
\alias{hispwrk}
\title{Does r know hispanics from work}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispwrk
}
\details{
Question Now, I’m going to ask some questions about these people you personally know.
D. Do you know any of these Hispanics from the place where you work?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2353/vshow}{GSS Data Explorer page for \code{hispwrk}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2142 \tab 420 \tab 206 \tab 1 \tab 48 \tab 2817 \cr
   Total \tab 2142 \tab 420 \tab 206 \tab 1 \tab 48 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{mostcom}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
