% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{hit}
\alias{hit}
\title{Ever punched or beaten by another person}
\source{
General Social Survey https://gss.norc.org
}
\description{
hit
}
\details{
Question Have you ever been punched or beaten by another person?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/658/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab yes \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1 \tab 1090 \tab 413 \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab 1017 \tab 472 \tab 1 \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab 1075 \tab 424 \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 2 \tab 993 \tab 537 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab 977 \tab 490 \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1 \tab 868 \tab 729 \tab 1 \tab - \tab 1599 \cr
   1984 \tab - \tab 1 \tab 885 \tab 585 \tab 2 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab 935 \tab 535 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab 1161 \tab 653 \tab 5 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 2 \tab 643 \tab 352 \tab - \tab - \tab 1481 \cr
   1989 \tab 502 \tab - \tab 657 \tab 378 \tab - \tab - \tab 1537 \cr
   1990 \tab 473 \tab 1 \tab 568 \tab 329 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 500 \tab - \tab 665 \tab 347 \tab 5 \tab - \tab 1517 \cr
   1993 \tab 549 \tab - \tab 677 \tab 379 \tab 1 \tab - \tab 1606 \cr
   1994 \tab 2481 \tab - \tab 315 \tab 195 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 49476 \tab 8 \tab 12526 \tab 6818 \tab 18 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{antiwar}},
\code{\link{civright}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{hitage}},
\code{\link{hitnum}},
\code{\link{prowar}},
\code{\link{school}},
\code{\link{strike}}

Other Violence: 
\code{\link{hitage}},
\code{\link{hitnum}}
}
\concept{Core}
\concept{Violence}
\keyword{variable}
