% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_75.R
\name{hitbeatr}
\alias{hitbeatr}
\title{Stranger was beating up a woman}
\source{
General Social Survey https://gss.norc.org
}
\description{
hitbeatr
}
\details{
Question Are there any situations that you can imagine in which you would approve of a man punching an adult male stranger?
D. Was beating up a woman and the man saw it?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/668/vshow}{GSS Data Explorer page for \code{hitbeatr}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab 489 \tab 855 \tab 127 \tab 23 \tab 10 \tab 1504 \cr
   1975 \tab 407 \tab 862 \tab 157 \tab 62 \tab 2 \tab 1490 \cr
   1976 \tab - \tab 1178 \tab 256 \tab 56 \tab 9 \tab 1499 \cr
   1978 \tab - \tab 1223 \tab 247 \tab 59 \tab 3 \tab 1532 \cr
   1980 \tab - \tab 1196 \tab 221 \tab 43 \tab 8 \tab 1468 \cr
   1983 \tab - \tab 1318 \tab 216 \tab 60 \tab 5 \tab 1599 \cr
   1984 \tab - \tab 1201 \tab 203 \tab 54 \tab 15 \tab 1473 \cr
   1986 \tab - \tab 1187 \tab 232 \tab 43 \tab 8 \tab 1470 \cr
   1987 \tab - \tab 1431 \tab 282 \tab 94 \tab 12 \tab 1819 \cr
   1988 \tab 484 \tab 816 \tab 131 \tab 43 \tab 7 \tab 1481 \cr
   1989 \tab 502 \tab 841 \tab 145 \tab 44 \tab 5 \tab 1537 \cr
   1990 \tab 473 \tab 750 \tab 117 \tab 30 \tab 2 \tab 1372 \cr
   1991 \tab 500 \tab 815 \tab 158 \tab 37 \tab 7 \tab 1517 \cr
   1993 \tab 549 \tab 888 \tab 130 \tab 36 \tab 3 \tab 1606 \cr
   1994 \tab 2481 \tab 415 \tab 78 \tab 16 \tab 2 \tab 2992 \cr
   Total \tab 5885 \tab 14976 \tab 2700 \tab 700 \tab 98 \tab 24359 \cr
}
}

\seealso{
Other Core: 
\code{\link{gunnum}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{polabuse}},
\code{\link{polhitok}},
\code{\link{polmurdr}}
}
\concept{Core}
\keyword{variable}
