% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_75.R
\name{hitok}
\alias{hitok}
\title{Ever approve of man punching adult male}
\source{
General Social Survey https://gss.norc.org
}
\description{
hitok
}
\details{
Question Are there any situations that you can imagine in which you would approve of a man punching an adult male stranger?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/664/vshow}{GSS Data Explorer page for \code{hitok}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 982 \tab 489 \tab 33 \tab - \tab 1504 \cr
   1975 \tab - \tab 1023 \tab 407 \tab 49 \tab 11 \tab 1490 \cr
   1976 \tab - \tab 985 \tab 439 \tab 67 \tab 8 \tab 1499 \cr
   1978 \tab - \tab 960 \tab 493 \tab 71 \tab 8 \tab 1532 \cr
   1980 \tab - \tab 914 \tab 492 \tab 45 \tab 17 \tab 1468 \cr
   1983 \tab - \tab 1087 \tab 446 \tab 64 \tab 2 \tab 1599 \cr
   1984 \tab - \tab 875 \tab 533 \tab 52 \tab 13 \tab 1473 \cr
   1986 \tab - \tab 922 \tab 465 \tab 62 \tab 21 \tab 1470 \cr
   1987 \tab - \tab 1080 \tab 610 \tab 113 \tab 16 \tab 1819 \cr
   1988 \tab 484 \tab 617 \tab 322 \tab 45 \tab 13 \tab 1481 \cr
   1989 \tab 502 \tab 615 \tab 326 \tab 80 \tab 14 \tab 1537 \cr
   1990 \tab 473 \tab 562 \tab 277 \tab 52 \tab 8 \tab 1372 \cr
   1991 \tab 500 \tab 554 \tab 386 \tab 65 \tab 12 \tab 1517 \cr
   1993 \tab 549 \tab 634 \tab 354 \tab 63 \tab 6 \tab 1606 \cr
   1994 \tab 2481 \tab 313 \tab 175 \tab 21 \tab 2 \tab 2992 \cr
   Total \tab 4989 \tab 12123 \tab 6214 \tab 882 \tab 151 \tab 24359 \cr
}
}

\seealso{
Other Core: 
\code{\link{gunnum}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitrobbr}},
\code{\link{polabuse}},
\code{\link{polhitok}},
\code{\link{polmurdr}}
}
\concept{Core}
\keyword{variable}
