% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{hivdscrm}
\alias{hivdscrm}
\title{There is a lot of discrimination against people with hiv}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivdscrm
}
\details{
Question (Please indicate how much you agree or disagree with each of the following comments.) There is a lot of discrimination against people with HIV in this country today.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly disagree
\item \verb{[2]} somewhat disagree
\item \verb{[3]} somewhat agree
\item \verb{[4]} strongly agree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7558/vshow}{GSS Data Explorer page for \code{hivdscrm}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly disagree \tab somewhat disagree \tab somewhat agree \tab strongly agree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1714 \tab 161 \tab 490 \tab 945 \tab 380 \tab 316 \tab 4 \tab 22 \tab 4032 \cr
   2022 \tab 1929 \tab 113 \tab 321 \tab 636 \tab 341 \tab 172 \tab 2 \tab 30 \tab 3544 \cr
   2024 \tab 1188 \tab 161 \tab 483 \tab 780 \tab 368 \tab 254 \tab 8 \tab 67 \tab 3309 \cr
   Total \tab 4831 \tab 435 \tab 1294 \tab 2361 \tab 1089 \tab 742 \tab 14 \tab 119 \tab 10885 \cr
}
}

\keyword{variable}
