% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{hivimmrl}
\alias{hivimmrl}
\title{People with hiv have participated in immoral activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivimmrl
}
\details{
Question (Please indicate how much you agree or disagree with each of the following comments.) People who have HIV have participated in immoral activities.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly disagree
\item \verb{[2]} somewhat disagree
\item \verb{[3]} somewhat agree
\item \verb{[4]} strongly agree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7557/vshow}{GSS Data Explorer page for \code{hivimmrl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly disagree \tab somewhat disagree \tab somewhat agree \tab strongly agree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1713 \tab 1163 \tab 500 \tab 280 \tab 89 \tab 254 \tab 6 \tab 27 \tab 4032 \cr
   2022 \tab 1929 \tab 761 \tab 336 \tab 229 \tab 82 \tab 177 \tab 2 \tab 28 \tab 3544 \cr
   2024 \tab 1188 \tab 865 \tab 454 \tab 341 \tab 120 \tab 264 \tab 10 \tab 67 \tab 3309 \cr
   Total \tab 4830 \tab 2789 \tab 1290 \tab 850 \tab 291 \tab 695 \tab 18 \tab 122 \tab 10885 \cr
}
}

\keyword{variable}
