% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{hivkiss}
\alias{hivkiss}
\title{Kiss can spread hiv}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivkiss
}
\details{
Question D. People are likely to get HIV by deep kissing, putting their tongue in their partner's mouth, if their partner has HIV.
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5080/vshow}{GSS Data Explorer page for \code{hivkiss}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2014 \tab 707 \tab 481 \tab 1210 \tab 102 \tab 38 \tab - \tab 2538 \cr
   Total \tab 707 \tab 481 \tab 1210 \tab 102 \tab 38 \tab 0 \tab 2538 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{frndsex}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{idu30}},
\code{\link{matesex}},
\code{\link{nocondom}},
\code{\link{nodocidu}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexornt}},
\code{\link{sexornt1}},
\code{\link{sexpaid}},
\code{\link{sexsex}},
\code{\link{sexsex5}}

Other Aids: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidscndm}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidslook}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{hivvac}}

Other Health: 
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{anscitst}},
\code{\link{anydoc}},
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{crelesse}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{diffcare}},
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld4}},
\code{\link{disabld5}},
\code{\link{disabld6}},
\code{\link{disabld7}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc15a}},
\code{\link{doc16}},
\code{\link{doc16a}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc18a}},
\code{\link{doc19}},
\code{\link{doc19a}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc20a}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{docaskme}},
\code{\link{doccosts}},
\code{\link{docdecid}},
\code{\link{docinfo}},
\code{\link{doclist}},
\code{\link{docnomri}},
\code{\link{docrely}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{docvig}},
\code{\link{downblue}},
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downfeel}},
\code{\link{downqol}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotfeel}},
\code{\link{emotqol}},
\code{\link{emotrely}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{energy}},
\code{\link{findnwjb}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{health}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{helpsick}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hlthall}},
\code{\link{hlthcare}},
\code{\link{hlthcovr}},
\code{\link{hlthdays}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthinsr}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthplan}},
\code{\link{hlthtv}},
\code{\link{hlthtype}},
\code{\link{hlthwww}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hrdshp6}},
\code{\link{hyperten}},
\code{\link{insrchng}},
\code{\link{insrlmts}},
\code{\link{intmed}},
\code{\link{limitedp}},
\code{\link{misswork}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{modact}},
\code{\link{mykids}},
\code{\link{nodocidu}},
\code{\link{othkids}},
\code{\link{othplan}},
\code{\link{pain}},
\code{\link{painarms}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}},
\code{\link{peaceful}},
\code{\link{physhlth}},
\code{\link{regdoc}},
\code{\link{regdocyr}},
\code{\link{sathealt}},
\code{\link{secopin}},
\code{\link{sexpaid}},
\code{\link{slpprblm}},
\code{\link{socacts}},
\code{\link{sppart}},
\code{\link{stairs}},
\code{\link{swithlth}},
\code{\link{treat1}},
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{usedmntl}},
\code{\link{usedup}},
\code{\link{wait1}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}},
\code{\link{weight}}
}
\concept{Aids}
\concept{Health}
\concept{High Risk Behaviors}
\keyword{variable}
