% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{hivtest1}
\alias{hivtest1}
\title{In what month and year was your last hiv test}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivtest1
}
\details{
Question Have you ever been tested for HIV? Do not count tests you may have had as part of a blood donation. Include oral test (where they take a swab from your mouth). A. Not including blood donations, in what month and year was your last HIV test?
As of 2024, HIVTEST1 accounts for only the year in which the respondent tested for HIV and the variable has been slightly adjusted to help protect confidentiality. Please see the 2024 codebook for more information. The original HIVTEST1 containing data from 2006 to 2022 has been renamed HIVTEST1_0622.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022, 2024 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{hivtest}},
\code{\link{hivtest2}},
\code{\link{idu30}},
\code{\link{relatsex}}

Other Health: 
\code{\link{aidslook}},
\code{\link{hivtest}},
\code{\link{hivtest2}}
}
\concept{Health}
\concept{High Risk Behaviors}
\keyword{variable}
