% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{hivtest2}
\alias{hivtest2}
\title{Where did you have your last hiv test}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivtest2
}
\details{
Question Have you ever been tested for HIV? Do not count tests you may have had as part of a blood donation. Include oral test (where they take a swab from your mouth). B. Where did you have your last HIV test – at a private doctor or HMO office, at a counselling and testing site, at a hospital, at a clinic, in a jail or prison, at home, or somewhere else?
As of 2024, HIVTEST2 has been recoded to only account for seven categories by combining jail, home, and drug treatment facility into "Somewhere else." This has been retroactively applied to all previous years of HIVTEST2. The original HIVTEST2 containing data from 2006 to 2022 has been renamed HIVTEST2_0622.
}
\section{Values}{

\itemize{
\item \code{1} private doctor or hmo office
\item \code{2} counseling and testing site
\item \code{3} hospital
\item \code{4} clinic
\item \code{5} jail or prison
\item \code{6} home
\item \code{7} somewhere else
\item \code{8} emergency room
\item \code{9} drug treatment facility
\item \code{10} hospital inpatient
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5076/vshow}{GSS Data Explorer page for \code{hivtest2}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab private doctor or hmo office \tab counseling and testing site \tab hospital \tab clinic \tab jail or prison \tab home \tab somewhere else \tab don't know \tab no answer \tab emergency room \tab drug treatment facility \tab hospital inpatient \tab skipped on web \tab Total \cr
   2006 \tab 3556 \tab 458 \tab 48 \tab 145 \tab 185 \tab 19 \tab 28 \tab 48 \tab 5 \tab 18 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1295 \tab 325 \tab 41 \tab 103 \tab 162 \tab 13 \tab 14 \tab 50 \tab 3 \tab 17 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1300 \tab 342 \tab 46 \tab 108 \tab 149 \tab 18 \tab 17 \tab 38 \tab 2 \tab 24 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1249 \tab 343 \tab 25 \tab 106 \tab 159 \tab 11 \tab 9 \tab 39 \tab 1 \tab 32 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1510 \tab 456 \tab 35 \tab 187 \tab 227 \tab 19 \tab 17 \tab 46 \tab 3 \tab 38 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2135 \tab 312 \tab 39 \tab 126 \tab 174 \tab 11 \tab 8 \tab 35 \tab 1 \tab 26 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1768 \tab 290 \tab 16 \tab 94 \tab 119 \tab 10 \tab 6 \tab 21 \tab - \tab 24 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3294 \tab 418 \tab 43 \tab - \tab 123 \tab 8 \tab 9 \tab 55 \tab - \tab 21 \tab 17 \tab 9 \tab 35 \tab - \tab 4032 \cr
   2022 \tab 2947 \tab 284 \tab 31 \tab 19 \tab 31 \tab 113 \tab 13 \tab 6 \tab 23 \tab 34 \tab 4 \tab 38 \tab - \tab 1 \tab 3544 \cr
   Total \tab 19054 \tab 3228 \tab 324 \tab 888 \tab 1329 \tab 222 \tab 121 \tab 338 \tab 38 \tab 234 \tab 21 \tab 47 \tab 35 \tab 1 \tab 25880 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{idu30}},
\code{\link{relatsex}}
}
\concept{High Risk Behaviors}
\keyword{variable}
