% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_463.R
\name{hlpdown}
\alias{hlpdown}
\title{Who r would first turn to if they needed help because they were feeling down or}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlpdown
}
\details{
Question Who would you turn to first to~. Be there for you if you felt a bit down or depressed and wanted to talk about it?
}
\section{Values}{

\itemize{
\item \verb{[1]} close family member
\item \verb{[2]} more distant family member
\item \verb{[3]} close friend
\item \verb{[4]} neighbor
\item \verb{[5]} someone i work with
\item \verb{[6]} someone else
\item \verb{[7]} no one
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7246/vshow}{GSS Data Explorer page for \code{hlpdown}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab close family member \tab more distant family member \tab close friend \tab neighbor \tab someone i work with \tab someone else \tab no one \tab don't know \tab Total \cr
   2018 \tab 1175 \tab 578 \tab 21 \tab 491 \tab 6 \tab 9 \tab 20 \tab 43 \tab 5 \tab 2348 \cr
   Total \tab 1175 \tab 578 \tab 21 \tab 491 \tab 6 \tab 9 \tab 20 \tab 43 \tab 5 \tab 2348 \cr
}
}

\keyword{variable}
