% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{hlpequip}
\alias{hlpequip}
\title{Enough help and equip to ge the job done}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlpequip
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
C. I receive enough help and equipment to get the job done
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2798/vshow}{GSS Data Explorer page for \code{hlpequip}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 857 \tab 677 \tab 171 \tab 68 \tab 9 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 759 \tab 679 \tab 184 \tab 77 \tab 18 \tab 17 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 536 \tab 466 \tab 106 \tab 52 \tab 2 \tab 25 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 589 \tab 511 \tab 98 \tab 42 \tab 3 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 679 \tab 567 \tab 128 \tab 39 \tab 4 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1552 \tab 848 \tab 821 \tab 203 \tab 76 \tab 10 \tab 12 \tab 22 \tab 3544 \cr
   Total \tab 8342 \tab 4268 \tab 3721 \tab 890 \tab 354 \tab 46 \tab 106 \tab 22 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{cowrkint}},
\code{\link{fringeok}},
\code{\link{haveinfo}},
\code{\link{jobsecok}},
\code{\link{opdevel}},
\code{\link{promtefr}},
\code{\link{supcares}},
\code{\link{wkfreedm}}
}
\concept{Quality of Working Life}
\keyword{variable}
