% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{hlpinflu}
\alias{hlpinflu}
\title{Does r have people to ask to influence important decision}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlpinflu
}
\details{
Question Some people because of their job, position in the community or contacts, are asked by others to  help influence important decisions in their favor.
B. And are there people you could ask to help influence important decisions in your favor?
}
\section{Values}{

\itemize{
\item \code{1} no, nobody
\item \code{2} yes, a few people
\item \code{3} yes, some people
\item \code{4} yes, a lot of people
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4120/vshow}{GSS Data Explorer page for \code{hlpinflu}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no, nobody \tab yes, a few people \tab yes, some people \tab yes, a lot of people \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 566 \tab 630 \tab 229 \tab 85 \tab 6 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 566 \tab 630 \tab 229 \tab 85 \tab 6 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{polsfair}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{wotrial}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
