% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{hlpneedy}
\alias{hlpneedy}
\title{Have ever contributed time to help the needy}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlpneedy
}
\details{
Question Over the past 5 years have you done any of the following to express your opinion about an issue or your support for a cause?
F. Contribute your time to help the needy
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2915/vshow}{GSS Data Explorer page for \code{hlpneedy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 690 \tab 690 \tab 8 \tab 5 \tab 2765 \cr
   Total \tab 1372 \tab 690 \tab 690 \tab 8 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{givchng}},
\code{\link{partthon}}
}
\concept{Adult Transitions}
\keyword{variable}
