% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{hlpsoc}
\alias{hlpsoc}
\title{Importance of social usefulness in a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlpsoc
}
\details{
Question On the following list there are various aspects of jobs. Please circle one number to show how important you personally consider it is in a job:
H. A job that is useful to society?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} neither important nor unimportant
\item \code{4} not important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4514/vshow}{GSS Data Explorer page for \code{hlpsoc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab neither important nor unimportant \tab not important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 406 \tab 690 \tab 248 \tab 59 \tab 7 \tab 17 \tab 26 \tab 1537 \cr
   1998 \tab 1609 \tab 362 \tab 552 \tab 207 \tab 34 \tab 7 \tab 20 \tab 41 \tab 2832 \cr
   2006 \tab 2992 \tab 697 \tab 659 \tab 121 \tab 26 \tab 8 \tab 5 \tab 2 \tab 4510 \cr
   2016 \tab 1390 \tab 697 \tab 596 \tab 141 \tab 26 \tab 9 \tab 3 \tab 5 \tab 2867 \cr
   Total \tab 6075 \tab 2162 \tab 2497 \tab 717 \tab 145 \tab 31 \tab 45 \tab 74 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
