% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_562.R
\name{hlth10}
\alias{hlth10}
\title{Respondent in hospital or sanatorium}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlth10
}
\details{
Question Now I am going to ask about some things that happen in families.
First I am going to ask you about illness.
10. During the past twelve months, that is since (CURRENT MONTH) 2003, have you, yourself, been a patient in a hospital, sanitorium, convalescent or nursing home (apart from having a baby)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5145/vshow}{GSS Data Explorer page for \code{hlth10}}.
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth11}},
\code{\link{hlth12}},
\code{\link{hlth13}},
\code{\link{hlth14}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{work1}},
\code{\link{work2}},
\code{\link{work3}}

Other Health: 
\code{\link{hlth11}},
\code{\link{hlth8}},
\code{\link{hlth9}}

Other Trauma: 
\code{\link{hlth11}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{work1}},
\code{\link{work2}},
\code{\link{work3}}
}
\concept{Health}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
