% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_561.R
\name{hlth3}
\alias{hlth3}
\title{Infertility, unable to have a baby}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlth3
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
A. First, thinking about health related matters, did any of the following happen to you since (CURRENT MONTH, 1990/2003)?
3. Infertility or unable to have a baby.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5138/vshow}{GSS Data Explorer page for \code{hlth3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1991 \tab 500 \tab 35 \tab 975 \tab 7 \tab - \tab 1517 \cr
   2004 \tab 1472 \tab 26 \tab 1304 \tab 9 \tab 1 \tab 2812 \cr
   Total \tab 1972 \tab 61 \tab 2279 \tab 16 \tab 1 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth1}},
\code{\link{hlth2}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}}
}
\concept{Negative Life Events}
\keyword{variable}
