% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_561.R
\name{hlth5}
\alias{hlth5}
\title{Illegal drugs (marijuana, cocaine)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlth5
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific vents, please think carefully, so I can record things accurately.
A. First, thinking about health related matters, did any of the following happen to you since (CURRENT MONTH, 1990/2003)?
5. Used illegal drugs (e.g. marijuana, cocaine, pills).
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5140/vshow}{GSS Data Explorer page for \code{hlth5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 500 \tab 30 \tab 982 \tab 5 \tab 1517 \cr
   2004 \tab 1472 \tab 77 \tab 1253 \tab 10 \tab 2812 \cr
   Total \tab 1972 \tab 107 \tab 2235 \tab 15 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth1}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}}
}
\concept{Negative Life Events}
\keyword{variable}
