% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_561.R
\name{hlth6}
\alias{hlth6}
\title{(Husband, wife)partner in hospital}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlth6
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
A. First, thinking about health related matters, did any of the following happen to you since (CURRENT MONTH, 1990/2003)?
6. Your (husband/wife)/partner was a patient in a hospital, sanitorium, convalescent or nursing home
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5141/vshow}{GSS Data Explorer page for \code{hlth6}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1991 \tab 500 \tab 73 \tab 927 \tab 2 \tab 15 \tab 1517 \cr
   2004 \tab 1472 \tab 135 \tab 1193 \tab 3 \tab 9 \tab 2812 \cr
   Total \tab 1972 \tab 208 \tab 2120 \tab 5 \tab 24 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth1}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth7}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}}
}
\concept{Negative Life Events}
\keyword{variable}
