% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_647.R
\name{hlthacc2}
\alias{hlthacc2}
\title{Old vs young access to healthcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthacc2
}
\details{
Question In the United States, do you think it is easier or harder to get access to health care... for old people than for young people
}
\section{Values}{

\itemize{
\item \code{1} much easier
\item \code{2} somewhat easier
\item \code{3} about the same
\item \code{4} somewhat harder
\item \code{5} much harder
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7798/vshow}{GSS Data Explorer page for \code{hlthacc2}}.

Counts by year:\tabular{llllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab Total \cr
   2022 \tab 2458 \tab 90 \tab 217 \tab 404 \tab 247 \tab 128 \tab 3544 \cr
   Total \tab 2458 \tab 90 \tab 217 \tab 404 \tab 247 \tab 128 \tab 3544 \cr
}
}

\keyword{variable}
