% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_647.R
\name{hlthacc3}
\alias{hlthacc3}
\title{Men vs women access to healthcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthacc3
}
\details{
Question In the United States, do you think it is easier or harder to get access to health care... for women than for men
}
\section{Values}{

\itemize{
\item \verb{[1]} much easier
\item \verb{[2]} somewhat easier
\item \verb{[3]} about the same
\item \verb{[4]} somewhat harder
\item \verb{[5]} much harder
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7799/vshow}{GSS Data Explorer page for \code{hlthacc3}}.

Counts by year:\tabular{llllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab Total \cr
   2022 \tab 2485 \tab 47 \tab 87 \tab 721 \tab 154 \tab 50 \tab 3544 \cr
   Total \tab 2485 \tab 47 \tab 87 \tab 721 \tab 154 \tab 50 \tab 3544 \cr
}
}

\keyword{variable}
