% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{hlthbtr}
\alias{hlthbtr}
\title{Higher incomes afford better health care}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthbtr
}
\details{
Question Is it fair or unfair that people with higher incomes can afford better health care than people with lower incomes? Would you say it is...
}
\section{Values}{

\itemize{
\item \verb{[1]} very fair
\item \verb{[2]} somewhat fair
\item \verb{[3]} neither fair nor unfair
\item \verb{[4]} somewhat unfair
\item \verb{[5]} very unfair
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7790/vshow}{GSS Data Explorer page for \code{hlthbtr}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very fair \tab somewhat fair \tab neither fair nor unfair \tab somewhat unfair \tab very unfair \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2378 \tab 68 \tab 161 \tab 226 \tab 223 \tab 439 \tab 38 \tab 3 \tab 8 \tab 3544 \cr
   Total \tab 2378 \tab 68 \tab 161 \tab 226 \tab 223 \tab 439 \tab 38 \tab 3 \tab 8 \tab 3544 \cr
}
}

\keyword{variable}
