% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{hlthctzn}
\alias{hlthctzn}
\title{Access to public funded health care if not citizen}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthctzn
}
\details{
Question How much do you agree or disagree with the following statements? People should have access to publicly funded health care even if they... Do not hold citizenship of the United States.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7795/vshow}{GSS Data Explorer page for \code{hlthctzn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2022 \tab 2379 \tab 169 \tab 331 \tab 184 \tab 231 \tab 205 \tab 44 \tab 1 \tab 3544 \cr
   Total \tab 2379 \tab 169 \tab 331 \tab 184 \tab 231 \tab 205 \tab 44 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
