% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{hlthdmg}
\alias{hlthdmg}
\title{Access to public funded health care if damage own health}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthdmg
}
\details{
Question (People should have access to publicly funded health care even if they...) Behave in ways that damage their health.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7796/vshow}{GSS Data Explorer page for \code{hlthdmg}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2379 \tab 130 \tab 386 \tab 263 \tab 232 \tab 87 \tab 57 \tab 5 \tab 5 \tab 3544 \cr
   Total \tab 2379 \tab 130 \tab 386 \tab 263 \tab 232 \tab 87 \tab 57 \tab 5 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
