% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_277.R
\name{hlthfrel}
\alias{hlthfrel}
\title{R sought health information friends or relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthfrel
}
\details{
Question If YES, please tell me if you tried to find such health information from:
E. Friends or relatives?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2521/vshow}{GSS Data Explorer page for \code{hlthfrel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 2412 \tab 201 \tab 122 \tab 61 \tab 21 \tab - \tab - \tab 2817 \cr
   2002 \tab 2104 \tab 280 \tab 229 \tab 83 \tab 6 \tab 63 \tab - \tab 2765 \cr
   2004 \tab 2209 \tab 230 \tab 199 \tab 91 \tab 6 \tab 76 \tab 1 \tab 2812 \cr
   Total \tab 6725 \tab 711 \tab 550 \tab 235 \tab 33 \tab 139 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{hlthdoc}},
\code{\link{hlthmag2}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polpapr}},
\code{\link{poltv}}
}
\concept{Information Society}
\keyword{variable}
