% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_647.R
\name{hlthgene}
\alias{hlthgene}
\title{Suffer health prob because of genes}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthgene
}
\details{
Question (People suffer from severe health problems...) Because of their genes.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7803/vshow}{GSS Data Explorer page for \code{hlthgene}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2380 \tab 128 \tab 633 \tab 279 \tab 51 \tab 12 \tab 54 \tab 1 \tab 6 \tab 3544 \cr
   Total \tab 2380 \tab 128 \tab 633 \tab 279 \tab 51 \tab 12 \tab 54 \tab 1 \tab 6 \tab 3544 \cr
}
}

\keyword{variable}
