% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{hlthinf}
\alias{hlthinf}
\title{Health care system in us inefficient}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthinf
}
\details{
Question (How much do you agree or disagree with the following statements?) In general, the health care system in the United States is inefficient.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7793/vshow}{GSS Data Explorer page for \code{hlthinf}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2378 \tab 257 \tab 415 \tab 246 \tab 166 \tab 36 \tab 40 \tab 2 \tab 4 \tab 3544 \cr
   Total \tab 2378 \tab 257 \tab 415 \tab 246 \tab 166 \tab 36 \tab 40 \tab 2 \tab 4 \tab 3544 \cr
}
}

\keyword{variable}
