% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{hlthinfo}
\alias{hlthinfo}
\title{R sought health information past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthinfo
}
\details{
Question In the past year, that is, between (INSERT CURRENT MONTH) 1999/2001/2003 and (INSERT CURRENT MONTH) 2000/02"have you looked for information about a health concern or medical problem?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2516/vshow}{GSS Data Explorer page for \code{hlthinfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 1884 \tab 528 \tab 219 \tab 171 \tab 15 \tab - \tab - \tab 2817 \cr
   2002 \tab 1372 \tab 724 \tab 331 \tab 202 \tab 6 \tab 122 \tab 8 \tab 2765 \cr
   2004 \tab 1828 \tab 376 \tab 125 \tab 186 \tab 6 \tab 286 \tab 5 \tab 2812 \cr
   Total \tab 5084 \tab 1628 \tab 675 \tab 559 \tab 27 \tab 408 \tab 13 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
