% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_243.R
\name{hlthinsr}
\alias{hlthinsr}
\title{Have health insurance}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthinsr
}
\details{
Question First, are you, yourself, covered by health insurance, a government plan like Medicare or Medicaid, or some other plan that pays for your medical care?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes covered
\item \verb{[2]} no not covered
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2185/vshow}{GSS Data Explorer page for \code{hlthinsr}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes covered \tab no not covered \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 1189 \tab 193 \tab 1 \tab 4 \tab 2832 \cr
   Total \tab 1445 \tab 1189 \tab 193 \tab 1 \tab 4 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{mntlcare}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid3}},
\code{\link{psycpsyc}},
\code{\link{swithlth}},
\code{\link{usedmntl}}
}
\concept{Mental Health}
\keyword{variable}
