% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{hlthmag1}
\alias{hlthmag1}
\title{R sought health information general interest magazines}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthmag1
}
\details{
Question If YES, please tell me if you tried to find such health information from:
B. Articles in a general-interest magazine?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2518/vshow}{GSS Data Explorer page for \code{hlthmag1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab Total \cr
   2000 \tab 2412 \tab 242 \tab 92 \tab 50 \tab 21 \tab - \tab 2817 \cr
   2002 \tab 2104 \tab 446 \tab 121 \tab 57 \tab 6 \tab 31 \tab 2765 \cr
   2004 \tab 2209 \tab 382 \tab 127 \tab 38 \tab 6 \tab 50 \tab 2812 \cr
   Total \tab 6725 \tab 1070 \tab 340 \tab 145 \tab 33 \tab 81 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthinfo}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
