% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_649.R
\name{hlthnot}
\alias{hlthnot}
\title{Felt couldn't overcome probs in last 4 weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthnot
}
\details{
Question (During the past 4 weeks how often...) Have you felt you could not overcome your problems?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} seldom
\item \verb{[3]} sometimes
\item \verb{[4]} often
\item \verb{[5]} very often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7824/vshow}{GSS Data Explorer page for \code{hlthnot}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab seldom \tab sometimes \tab often \tab very often \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2381 \tab 612 \tab 244 \tab 192 \tab 42 \tab 41 \tab 23 \tab 3 \tab 6 \tab 3544 \cr
   Total \tab 2381 \tab 612 \tab 244 \tab 192 \tab 42 \tab 41 \tab 23 \tab 3 \tab 6 \tab 3544 \cr
}
}

\keyword{variable}
