% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_649.R
\name{hlthpain}
\alias{hlthpain}
\title{Body aches or pains in last 4 weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthpain
}
\details{
Question (During the past 4 weeks how often...) Have you had bodily aches or pains?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} seldom
\item \verb{[3]} sometimes
\item \verb{[4]} often
\item \verb{[5]} very often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7821/vshow}{GSS Data Explorer page for \code{hlthpain}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab seldom \tab sometimes \tab often \tab very often \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2381 \tab 182 \tab 224 \tab 353 \tab 193 \tab 183 \tab 18 \tab 3 \tab 7 \tab 3544 \cr
   Total \tab 2381 \tab 182 \tab 224 \tab 353 \tab 193 \tab 183 \tab 18 \tab 3 \tab 7 \tab 3544 \cr
}
}

\keyword{variable}
