% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{hlthsat}
\alias{hlthsat}
\title{How satisfied r with health care system in us}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthsat
}
\details{
Question In general, how satisfied or dissatisfied are you with the health care system in the United States? Are you...
}
\section{Values}{

\itemize{
\item \code{1} completely satisfied
\item \code{2} very satisfied
\item \code{3} fairly satisfied
\item \code{4} neither satisfied nor dissatisfied
\item \code{5} fairly dissatisfied
\item \code{6} very dissatisfied
\item \code{7} completely dissatisfied
\item \code{8} does not apply
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7831/vshow}{GSS Data Explorer page for \code{hlthsat}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab completely satisfied \tab very satisfied \tab fairly satisfied \tab neither satisfied nor dissatisfied \tab fairly dissatisfied \tab very dissatisfied \tab completely dissatisfied \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2382 \tab 46 \tab 162 \tab 421 \tab 168 \tab 188 \tab 87 \tab 51 \tab 31 \tab 3 \tab 5 \tab 3544 \cr
   Total \tab 2382 \tab 46 \tab 162 \tab 421 \tab 168 \tab 188 \tab 87 \tab 51 \tab 31 \tab 3 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
