% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{hlthstrt}
\alias{hlthstrt}
\title{Would you say the respondent's health in general is excellent, good, fair, or poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthstrt
}
\details{
Question Would you say the respondent's health in general is Excellent, Good, Fair, or Poor?
}
\section{Values}{

\itemize{
\item \verb{[1]} excellent
\item \verb{[2]} good
\item \verb{[3]} fair
\item \verb{[4]} poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6651/vshow}{GSS Data Explorer page for \code{hlthstrt}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab excellent \tab good \tab fair \tab poor \tab not available in this year \tab Total \cr
   2016 \tab 1437 \tab 622 \tab 637 \tab 148 \tab 23 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 469 \tab 532 \tab 137 \tab 22 \tab - \tab 2348 \cr
   2022 \tab 2661 \tab 308 \tab 452 \tab 101 \tab 22 \tab - \tab 3544 \cr
   Total \tab 5286 \tab 1399 \tab 1621 \tab 386 \tab 67 \tab 0 \tab 8759 \cr
}
}

\keyword{variable}
