% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{hlthtype}
\alias{hlthtype}
\title{Source of health insurance}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthtype
}
\details{
Question What is the source of your health insurance?
}
\section{Values}{

\itemize{
\item \verb{[1]} my employer
\item \verb{[2]} employer of my spouse/partner
\item \verb{[3]} employer of someone else in my family
\item \verb{[4]} medicaid
\item \verb{[5]} medicare
\item \verb{[6]} individual plan from private insurer not related to current or past employment
\item \verb{[7]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3931/vshow}{GSS Data Explorer page for \code{hlthtype}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab my employer \tab employer of my spouse/partner \tab employer of someone else in my family \tab medicaid \tab medicare \tab individual plan from private insurer not related to current or past employment \tab other \tab no answer \tab Total \cr
   2008 \tab 1907 \tab 46 \tab 10 \tab 2 \tab 7 \tab 16 \tab 7 \tab 4 \tab 24 \tab 2023 \cr
   Total \tab 1907 \tab 46 \tab 10 \tab 2 \tab 7 \tab 16 \tab 7 \tab 4 \tab 24 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{hlthcovr}},
\code{\link{mykids}},
\code{\link{othkids}},
\code{\link{sppart}}
}
\concept{Sexual Orientation}
\keyword{variable}
