% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{hmemaker}
\alias{hmemaker}
\title{Full-time homemaker's lives improved}
\source{
General Social Survey https://gss.norc.org
}
\description{
hmemaker
}
\details{
Question We'd like your views on how the women's movement has affected certain groups. For each group I name, please tell me whether you think the women's movement has improved their lives, made their lives worse, or had no effect On their lives.
A. Full-time homemaker
}
\section{Values}{

\itemize{
\item \code{1} improved their life
\item \code{2} made their life worse
\item \code{3} had no effect on their life
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1700/vshow}{GSS Data Explorer page for \code{hmemaker}}.
}

\seealso{
Other Gender: 
\code{\link{fehire}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{mebear}},
\code{\link{wrkclass}}
}
\concept{Gender}
\keyword{variable}
