% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_244.R
\name{hmo2}
\alias{hmo2}
\title{Hmo prevent doctors from prescribing tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
hmo2
}
\details{
Question The health care system is changing rapidly with more and more people being treated in Health Maintenance organizations (or HMOs) and in managed care plans. How much do you agree or disagree with how the HMOs and managed care plans are working?
B. They prevent doctors from prescribing tests necessary for treatment.
}
\section{Values}{

\itemize{
\item \code{1} strongly disagree
\item \code{2} disagree
\item \code{3} neither agree nor disagree
\item \code{4} agree
\item \code{5} strongly agree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2194/vshow}{GSS Data Explorer page for \code{hmo2}}.
}

\seealso{
Other Mental Health: 
\code{\link{doc1}},
\code{\link{hmo1}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{mntldeny}},
\code{\link{mntldif}}
}
\concept{Mental Health}
\keyword{variable}
