% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_244.R
\name{hmo3}
\alias{hmo3}
\title{Hmo help to control costs}
\source{
General Social Survey https://gss.norc.org
}
\description{
hmo3
}
\details{
Question The health care system is changing rapidly with more and more people being treated in Health Maintenance organizations (or HMOs) and in managed care plans. How much do you agree or disagree with how the HMOs and managed care plans are working?
C. They help to control costs.
}
\section{Values}{

\itemize{
\item \code{1} strongly disagree
\item \code{2} disagree
\item \code{3} neither agree nor disagree
\item \code{4} agree
\item \code{5} strongly agree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2195/vshow}{GSS Data Explorer page for \code{hmo3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly disagree \tab disagree \tab neither agree nor disagree \tab agree \tab strongly agree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 53 \tab 253 \tab 204 \tab 584 \tab 83 \tab 206 \tab 4 \tab 2832 \cr
   Total \tab 1445 \tab 53 \tab 253 \tab 204 \tab 584 \tab 83 \tab 206 \tab 4 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc1}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{mntldeny}},
\code{\link{mntldif}}
}
\concept{Mental Health}
\keyword{variable}
