% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{homeband}
\alias{homeband}
\title{How home is connected to www}
\source{
General Social Survey https://gss.norc.org
}
\description{
homeband
}
\details{
Question Is your home connected to the World Wide Web by a telephone line, a DSL, cable, some other way, or not at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} telephone line
\item \verb{[2]} dsl
\item \verb{[3]} cable
\item \verb{[4]} satellite
\item \verb{[5]} wireless
\item \verb{[6]} cellphone
\item \verb{[7]} other
\item \verb{[8]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2656/vshow}{GSS Data Explorer page for \code{homeband}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab telephone line \tab dsl \tab cable \tab satellite \tab wireless \tab cellphone \tab other \tab not at all \tab don't know \tab Total \cr
   2004 \tab 992 \tab 785 \tab 385 \tab 400 \tab 5 \tab 2 \tab 2 \tab 4 \tab 213 \tab 24 \tab 2812 \cr
   Total \tab 992 \tab 785 \tab 385 \tab 400 \tab 5 \tab 2 \tab 2 \tab 4 \tab 213 \tab 24 \tab 2812 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{huclean}},
\code{\link{knowkid3}},
\code{\link{knowkid4}}
}
\concept{Information Society}
\keyword{variable}
