% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{homekid}
\alias{homekid}
\title{Most women really want a home and kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
homekid
}
\details{
Question Do you agree or disagree ...
E. A job is alright, but what most women really want is a home and children.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4347/vshow}{GSS Data Explorer page for \code{homekid}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 108 \tab 354 \tab 358 \tab 359 \tab 164 \tab 44 \tab 27 \tab - \tab 1481 \cr
   1994 \tab 1545 \tab 101 \tab 359 \tab 327 \tab 428 \tab 179 \tab 38 \tab 15 \tab - \tab 2992 \cr
   2002 \tab 1594 \tab 143 \tab 307 \tab 335 \tab - \tab 354 \tab 29 \tab 3 \tab - \tab 2765 \cr
   2012 \tab 672 \tab 83 \tab 360 \tab 299 \tab 372 \tab 120 \tab 60 \tab 8 \tab - \tab 1974 \cr
   2022 \tab 2378 \tab 44 \tab 208 \tab 333 \tab 287 \tab 234 \tab 41 \tab 7 \tab 12 \tab 3544 \cr
   Total \tab 6256 \tab 479 \tab 1588 \tab 1652 \tab 1446 \tab 1051 \tab 212 \tab 60 \tab 12 \tab 12756 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famsuffr}},
\code{\link{fejobind}},
\code{\link{hapifwrk}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{kidsuffr}},
\code{\link{mawrkwrm}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
