% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{homeyear}
\alias{homeyear}
\title{Rmost recent home you have purchased.}
\source{
General Social Survey https://gss.norc.org
}
\description{
homeyear
}
\details{
Question Tell me about the most recent home you have purchased. When did you buy this home?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1813/vshow}{GSS Data Explorer page for \code{homeyear}}.
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
