% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{homochng}
\alias{homochng}
\title{Homosexuality: inherent or choice?}
\source{
General Social Survey https://gss.norc.org
}
\description{
homochng
}
\details{
Question Do you think being homosexual is something people choose to be, or do you think it is something they cannot change?
}
\section{Values}{

\itemize{
\item \code{1} something people choose to be
\item \code{2} something they cannot change
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/635/vshow}{GSS Data Explorer page for \code{homochng}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab something people choose to be \tab something they cannot change \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2034 \tab 393 \tab 424 \tab 129 \tab 12 \tab 2992 \cr
   Total \tab 2034 \tab 393 \tab 424 \tab 129 \tab 12 \tab 2992 \cr
}
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{homosex}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{spdue}},
\code{\link{sppaid}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\keyword{variable}
