% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_664.R
\name{hompop_exp}
\alias{hompop_exp}
\title{Total people in hh (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hompop_exp
}
\details{
Question Calculated variable: Total people in household according to roster
As of 2024, HOMPOP_EXP has been calculated to only account for 0 to 14+ people in the household. This has been retroactively applied to all previous years of HOMPOP_EXP. The original HOMPOP_EXP containing data from in 2022 has been renamed HOMPOP_EXP_22.
}
\section{Values}{

\itemize{
\item \verb{[14]} 14 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8001/vshow}{GSS Data Explorer page for \code{hompop_exp}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 11 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab iap \tab 10 \tab 14 or more \tab Total \cr
   2022 \tab - \tab 32 \tab 580 \tab 1 \tab 510 \tab 265 \tab 219 \tab 83 \tab 36 \tab 18 \tab 8 \tab 6 \tab 1786 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab 569 \tab - \tab 475 \tab 224 \tab 193 \tab 90 \tab 42 \tab 16 \tab 10 \tab 3 \tab 1684 \tab 1 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 32 \tab 1149 \tab 1 \tab 985 \tab 489 \tab 412 \tab 173 \tab 78 \tab 34 \tab 18 \tab 9 \tab 3470 \tab 1 \tab 2 \tab 6853 \cr
}
}

\concept{Core}
\concept{Household}
\keyword{variable}
