% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{hompoph}
\alias{hompoph}
\title{Total people in hh (household screening version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hompoph
}
\details{
Question Calculated variable: total people in household according to roster. (ROSTER ONLY ASKED FOR ADULTS)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8000/vshow}{GSS Data Explorer page for \code{hompoph}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 9 \tab no answer \tab Total \cr
   2022 \tab - \tab 1401 \tab 1590 \tab 382 \tab 133 \tab 25 \tab 5 \tab 1 \tab 3 \tab 4 \tab 3544 \cr
   Total \tab 0 \tab 1401 \tab 1590 \tab 382 \tab 133 \tab 25 \tab 5 \tab 1 \tab 3 \tab 4 \tab 3544 \cr
}
}

\keyword{variable}
