% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_55.R
\name{honesty}
\alias{honesty}
\title{Honest-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
honesty
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
3. That a child is honest
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 most desirable
\item \verb{[2]} 3 most desirable
\item \verb{[3]} not mentioned
\item \verb{[4]} 3 least desirable
\item \verb{[5]} 1 least desirable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980, 1984, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/489/vshow}{GSS Data Explorer page for \code{honesty}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab no answer \tab 1 least desirable \tab Total \cr
   1980 \tab 976 \tab 161 \tab 148 \tab 174 \tab 4 \tab 5 \tab - \tab 1468 \cr
   1984 \tab - \tab 444 \tab 416 \tab 554 \tab 18 \tab 35 \tab 6 \tab 1473 \cr
   1986 \tab 740 \tab 177 \tab 192 \tab 337 \tab 9 \tab 14 \tab 1 \tab 1470 \cr
   Total \tab 1716 \tab 782 \tab 756 \tab 1065 \tab 31 \tab 54 \tab 7 \tab 4411 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicably}},
\code{\link{cleany}},
\code{\link{controly}},
\code{\link{interest}},
\code{\link{judgmeny}},
\code{\link{mannersy}},
\code{\link{roley}},
\code{\link{studious}},
\code{\link{successy}}
}
\concept{Core}
\keyword{variable}
