% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_230.R
\name{hope1}
\alias{hope1}
\title{I could get out of a jam}
\source{
General Social Survey https://gss.norc.org
}
\description{
hope1
}
\details{
Question (HAND CARD G2)
Using the scale on the showcard, please select the response that best describes how you think about yourself right now. Please focus on yourself and what is going on in your life at this moment. Please answer each item as I read them according to the scale on the card.
If I should find myself in a jam, I could think of many ways to get out of it.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely false
\item \verb{[2]} mostly false
\item \verb{[3]} somewhat false
\item \verb{[4]} slightly false
\item \verb{[5]} slightly true
\item \verb{[6]} somewhat true
\item \verb{[7]} mostly true
\item \verb{[8]} definitely true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6203/vshow}{GSS Data Explorer page for \code{hope1}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab definitely false \tab mostly false \tab somewhat false \tab slightly false \tab slightly true \tab somewhat true \tab mostly true \tab definitely true \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1409 \tab 19 \tab 23 \tab 27 \tab 28 \tab 154 \tab 299 \tab 543 \tab 346 \tab 7 \tab 12 \tab 2867 \cr
   Total \tab 1409 \tab 19 \tab 23 \tab 27 \tab 28 \tab 154 \tab 299 \tab 543 \tab 346 \tab 7 \tab 12 \tab 2867 \cr
}
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope2}},
\code{\link{hope3}},
\code{\link{hope4}},
\code{\link{lotr1}},
\code{\link{lotr2}},
\code{\link{lotr3}},
\code{\link{lotr4}},
\code{\link{lotr5}},
\code{\link{lotr6}}
}
\concept{Hope and Optimism}
\keyword{variable}
