% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_230.R
\name{hope4}
\alias{hope4}
\title{I see myself as successful}
\source{
General Social Survey https://gss.norc.org
}
\description{
hope4
}
\details{
Question (HAND CARD G2)
(Using the scale on the showcard, please select the response that best describes how you think about yourself right now.)
Right now, I see myself as being pretty successful.
}
\section{Values}{

\itemize{
\item \code{1} definitely false
\item \code{2} mostly false
\item \code{3} somewhat false
\item \code{4} slightly false
\item \code{5} slightly true
\item \code{6} somewhat true
\item \code{7} mostly true
\item \code{8} definitely true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6206/vshow}{GSS Data Explorer page for \code{hope4}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab definitely false \tab mostly false \tab somewhat false \tab slightly false \tab slightly true \tab somewhat true \tab mostly true \tab definitely true \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1409 \tab 51 \tab 48 \tab 57 \tab 83 \tab 215 \tab 337 \tab 426 \tab 225 \tab 5 \tab 11 \tab 2867 \cr
   Total \tab 1409 \tab 51 \tab 48 \tab 57 \tab 83 \tab 215 \tab 337 \tab 426 \tab 225 \tab 5 \tab 11 \tab 2867 \cr
}
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope1}},
\code{\link{hope2}},
\code{\link{hope3}},
\code{\link{lotr1}},
\code{\link{lotr2}},
\code{\link{lotr3}},
\code{\link{lotr4}},
\code{\link{lotr5}},
\code{\link{lotr6}}
}
\concept{Hope and Optimism}
\keyword{variable}
