% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{hope5}
\alias{hope5}
\title{I can think of many ways to reach my current goals}
\source{
General Social Survey https://gss.norc.org
}
\description{
hope5
}
\details{
Question (HAND CARD G2)
(Using the scale on the showcard, please select the response that best describes how you think about yourself right now.)
I can think of many ways to reach my current goals.
}
\section{Values}{

\itemize{
\item \code{1} definitely false
\item \code{2} mostly false
\item \code{3} somewhat false
\item \code{4} slightly false
\item \code{5} slightly true
\item \code{6} somewhat true
\item \code{7} mostly true
\item \code{8} definitely true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6207/vshow}{GSS Data Explorer page for \code{hope5}}.
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope6}}
}
\concept{Hope and Optimism}
\keyword{variable}
