% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{hospice}
\alias{hospice}
\title{Choose hospice until patient dies}
\source{
General Social Survey https://gss.norc.org
}
\description{
hospice
}
\details{
Question How much do you agree of disagree with the following options for this person?  This person should...
D. Choose hospice services where the patient will be cared for until the patient dies naturally.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2258/vshow}{GSS Data Explorer page for \code{hospice}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 457 \tab 636 \tab 148 \tab 60 \tab 10 \tab 40 \tab 36 \tab 2832 \cr
   Total \tab 1445 \tab 457 \tab 636 \tab 148 \tab 60 \tab 10 \tab 40 \tab 36 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{painmed}},
\code{\link{rightref}},
\code{\link{termvig}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\keyword{variable}
