% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{hosrel4}
\alias{hosrel4}
\title{Rels hospitalized betw 5 & 1 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
hosrel4
}
\details{
Question B. Number of relatives hospitalized/disabled during previous four years.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/745/vshow}{GSS Data Explorer page for \code{hosrel4}}.
}

\seealso{
Other Core: 
\code{\link{chlddth}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{madeath}},
\code{\link{padeath}},
\code{\link{unrel1}},
\code{\link{unrel4}}

Other Relatives: 
\code{\link{death16}},
\code{\link{death5}},
\code{\link{hosrel1}},
\code{\link{unrel1}},
\code{\link{unrel4}}

Other Trauma: 
\code{\link{death16}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{unrel1}},
\code{\link{unrel4}}
}
\concept{Core}
\concept{Relatives}
\concept{Trauma}
\keyword{variable}
