% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{howangry}
\alias{howangry}
\title{How intense would you say your anger was}
\source{
General Social Survey https://gss.norc.org
}
\description{
howangry
}
\details{
Question How intense would you say your anger or irritation was, if 0 means the least intense emotion that you can imagine experiencing and 10 is the most intense emotion that you can imagine having.
}
\section{Values}{

\itemize{
\item \code{0} weakest
\item \code{10} most intense
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1764/vshow}{GSS Data Explorer page for \code{howangry}}.
}

\seealso{
Other Emotions: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{thnkangr}}

Other Feelings: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{thnkangr}}
}
\concept{Emotions}
\concept{Feelings}
\keyword{variable}
