% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_564.R
\name{hrdshp3}
\alias{hrdshp3}
\title{Having the utilities shut-off}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrdshp3
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
D. Now I am going to ask about specific hardships. Did any of the following occur to you since (CURRENT MONTH), (1990/2003)?
3. Had your utilities (water, heat, or electricity) shut off.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5167/vshow}{GSS Data Explorer page for \code{hrdshp3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 500 \tab 29 \tab 984 \tab 4 \tab 1517 \cr
   2004 \tab 1472 \tab 56 \tab 1276 \tab 8 \tab 2812 \cr
   Total \tab 1972 \tab 85 \tab 2260 \tab 12 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{famper1}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}}
}
\concept{Negative Life Events}
\keyword{variable}
