% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_564.R
\name{hrdshp6}
\alias{hrdshp6}
\title{Lacking health insurance coverage}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrdshp6
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
D. Now I am going to ask about specific hardships. Did any of the following occur to you since (CURRENT MONTH), (1990/2003)?
6. Lacked health insurance coverage (e.g. Medicare, Medicaid, Blue Cross, an HMO, etc.).
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5170/vshow}{GSS Data Explorer page for \code{hrdshp6}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 500 \tab 122 \tab 888 \tab 7 \tab 1517 \cr
   2004 \tab 1472 \tab 246 \tab 1086 \tab 8 \tab 2812 \cr
   Total \tab 1972 \tab 368 \tab 1974 \tab 15 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{famper1}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp7}}
}
\concept{Negative Life Events}
\keyword{variable}
