% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{hrfambiz}
\alias{hrfambiz}
\title{How many hours per week?}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrfambiz
}
\details{
Question How many hours per week?
}
\section{Values}{

\itemize{
\item \verb{[1]} only especially busy seasons
\item \verb{[2]} less than 10 hours a week
\item \verb{[3]} between 10 and 30 hours a week
\item \verb{[4]} more than 30 hours a week
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1877/vshow}{GSS Data Explorer page for \code{hrfambiz}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab only especially busy seasons \tab less than 10 hours a week \tab between 10 and 30 hours a week \tab more than 30 hours a week \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 2841 \tab 8 \tab 7 \tab 12 \tab 27 \tab 9 \tab - \tab 2904 \cr
   Total \tab 2841 \tab 8 \tab 7 \tab 12 \tab 27 \tab 9 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{borrowed}},
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{econsys}},
\code{\link{evbuyhme}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{fixer1}},
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{fixhome}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{haggle}},
\code{\link{homeyear}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal3}},
\code{\link{inequal5}},
\code{\link{lawyer}},
\code{\link{lawyer1}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{layoffs}},
\code{\link{learnhme}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{natrecon}},
\code{\link{newowned}},
\code{\link{newused}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{realtor1}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{reldealr}},
\code{\link{relfixer}},
\code{\link{relhome}},
\code{\link{rellaw}},
\code{\link{relprivt}},
\code{\link{reqinfo}},
\code{\link{satcar}},
\code{\link{satfixes}},
\code{\link{sathome}},
\code{\link{satlawyr}},
\code{\link{sellauto}},
\code{\link{sellbaby}},
\code{\link{sellbed}},
\code{\link{sellhome}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{tagsales}},
\code{\link{transoth}},
\code{\link{transrel}},
\code{\link{typdealr}},
\code{\link{warrntyc}},
\code{\link{warrntyh}},
\code{\link{wholoand}},
\code{\link{whosold}},
\code{\link{wkfambiz}}

Other Labor Force Status: 
\code{\link{evwork}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{kdevwork}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawork14}},
\code{\link{mawrkgrw}},
\code{\link{neverwk}},
\code{\link{partfull}},
\code{\link{spevwork}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spwrksta}},
\code{\link{unemp}},
\code{\link{weekswrk}},
\code{\link{wkfambiz}},
\code{\link{wrkstat}}
}
\concept{Labor Force Status}
\concept{Markets}
\keyword{variable}
