% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_203.R
\name{hrseduc}
\alias{hrseduc}
\title{How many hours spent in education}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrseduc
}
\details{
Question I'd like your best estimate of the total number of hours you spent in the past month on each of the areas in which you have been a volunteer. First, how many hours did you spend in the past month working for (ASK FOR EACH ITEM R MENTIONED IN MONHLTH-MONOTH) (RECORD ABOVE UNDER MONHLTH-MONOTH TO NEAREST HALF HOUR. If UNSURE, PROBE FOR BEST ESTIMATE.)
B. Education
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1916/vshow}{GSS Data Explorer page for \code{hrseduc}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 12 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 22 \tab 25 \tab 3 \tab 30 \tab 4 \tab 40 \tab 42 \tab 45 \tab 48 \tab 5 \tab 6 \tab 7 \tab 8 \tab 80 \tab don't know \tab Total \cr
   1996 \tab 2779 \tab 4 \tab 14 \tab 3 \tab 2 \tab 4 \tab 1 \tab 24 \tab 7 \tab 1 \tab 3 \tab 11 \tab 1 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 9 \tab 8 \tab 1 \tab 7 \tab 1 \tab 1 \tab 2904 \cr
   Total \tab 2779 \tab 4 \tab 14 \tab 3 \tab 2 \tab 4 \tab 1 \tab 24 \tab 7 \tab 1 \tab 3 \tab 11 \tab 1 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 9 \tab 8 \tab 1 \tab 7 \tab 1 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{hrsenvir}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
