% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_203.R
\name{hrsenvir}
\alias{hrsenvir}
\title{How many hours spent in environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrsenvir
}
\details{
Question I'd like your best estimate of the total number of hours you spent in the past month on each of the areas in which you have been a volunteer. First, how many hours did you spend in the past month working for (ASK FOR EACH ITEM R MENTIONED IN MONHLTH-MONOTH) (RECORD ABOVE UNDER MONHLTH-MONOTH TO NEAREST HALF HOUR. If UNSURE, PROBE FOR BEST ESTIMATE.)
E. Environment
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1919/vshow}{GSS Data Explorer page for \code{hrsenvir}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 11 \tab 12 \tab 15 \tab 16 \tab 2 \tab 20 \tab 3 \tab 30 \tab 4 \tab 40 \tab 5 \tab 6 \tab 7 \tab 8 \tab Total \cr
   1996 \tab 2864 \tab 1 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2904 \cr
   Total \tab 2864 \tab 1 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{hrseduc}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
