% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_499.R
\name{hrsmoney}
\alias{hrsmoney}
\title{R preference re: work hrs and money}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrsmoney
}
\details{
Question Think of the number of hours you work and the money you earn in your main job, including regular overtime.If you had only one of these three choices, which of the following would you prefer?
}
\section{Values}{

\itemize{
\item \code{1} work longer hours and earn more money
\item \code{2} work the same number of hours and earn the same money
\item \code{3} work fewer hours and earn less money
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4532/vshow}{GSS Data Explorer page for \code{hrsmoney}}.
}

\seealso{
Other Work Orientation: 
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rsecjob}},
\code{\link{ryrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
