% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_204.R
\name{hrspol}
\alias{hrspol}
\title{How many hours spent in political org. etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrspol
}
\details{
Question I'd like your best estimate of the total number of hours you spent in the past month on each of the areas in which you have been a volunteer. First, how many hours did you spend in the past month working for (ASK FOR EACH ITEM R MENTIONED IN MONHLTH-MONOTH) (RECORD ABOVE UNDER MONHLTH-MONOTH TO NEAREST HALF HOUR. If UNSURE, PROBE FOR BEST ESTIMATE.)
J. Political organizations or campaigns
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1924/vshow}{GSS Data Explorer page for \code{hrspol}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 4 \tab 40 \tab 5 \tab 6 \tab 8 \tab 80 \tab Total \cr
   1996 \tab 2874 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2904 \cr
   Total \tab 2874 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{giveduc}},
\code{\link{givhlth}},
\code{\link{hrsart}},
\code{\link{hrsfound}},
\code{\link{hrsinfrm}},
\code{\link{hrsintl}},
\code{\link{hrsoth}},
\code{\link{hrswork}},
\code{\link{hrsyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
