% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{hsclass1}
\alias{hsclass1}
\title{R favorite subject in hs}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsclass1
}
\details{
Question B. When you were in high school/In high school), what (was/is) your favorite course or subject?
}
\section{Values}{

\itemize{
\item \code{0} none
\item \code{1} english, literature
\item \code{2} foreign language
\item \code{3} fine arts
\item \code{4} other humanities
\item \code{5} economics
\item \code{6} psychology
\item \code{7} government-civics
\item \code{8} social science-other social study
\item \code{9} mathematics
\item \code{10} science
\item \code{11} other
\item \code{12} business
\item \code{13} education
\item \code{14} engineering
\item \code{15} other vocational
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1452/vshow}{GSS Data Explorer page for \code{hsclass1}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab none \tab english, literature \tab foreign language \tab fine arts \tab other humanities \tab economics \tab psychology \tab government-civics \tab social science-other social study \tab mathematics \tab science \tab other \tab business \tab engineering \tab other vocational \tab no answer \tab Total \cr
   1993 \tab 139 \tab 141 \tab 297 \tab 27 \tab 112 \tab 25 \tab 23 \tab 15 \tab 51 \tab 134 \tab 295 \tab 119 \tab 144 \tab 50 \tab 1 \tab 26 \tab 7 \tab 1606 \cr
   Total \tab 139 \tab 141 \tab 297 \tab 27 \tab 112 \tab 25 \tab 23 \tab 15 \tab 51 \tab 134 \tab 295 \tab 119 \tab 144 \tab 50 \tab 1 \tab 26 \tab 7 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{hsclass2}},
\code{\link{mostimp4}}
}
\concept{Culture}
\keyword{variable}
