% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{hsjobtr}
\alias{hsjobtr}
\title{School subjects: job training}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsjobtr
}
\details{
Question Here are some things that might be taught in school. How important is it that schools teach each of these to 15-year-olds?
F. Job training.
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4020/vshow}{GSS Data Explorer page for \code{hsjobtr}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 273 \tab 237 \tab 112 \tab 29 \tab 10 \tab 6 \tab 10 \tab 1534 \cr
   Total \tab 857 \tab 273 \tab 237 \tab 112 \tab 29 \tab 10 \tab 6 \tab 10 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
