% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{hsorder}
\alias{hsorder}
\title{School subjects: discipline and order}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsorder
}
\details{
Question Here are some things that might be taught in school. How important is it that schools teach each of these to 15-year-olds?
I. Discipline and orderliness.
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4023/vshow}{GSS Data Explorer page for \code{hsorder}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 312 \tab 233 \tab 89 \tab 13 \tab 7 \tab 10 \tab 13 \tab 1534 \cr
   Total \tab 857 \tab 312 \tab 233 \tab 89 \tab 13 \tab 7 \tab 10 \tab 13 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hssci}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
