% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_133.R
\name{hsphouse}
\alias{hsphouse}
\title{Hispanics discriminated in housing}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsphouse
}
\details{
Question How much discrimination is there that makes it hard for Hispanic Americans to buy or rent housing wherever they want?
A. Hispanic Americans
}
\section{Values}{

\itemize{
\item \verb{[1]} a lot
\item \verb{[2]} some
\item \verb{[3]} only a little
\item \verb{[4]} none at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1246/vshow}{GSS Data Explorer page for \code{hsphouse}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a lot \tab some \tab only a little \tab none at all \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 251 \tab 572 \tab 273 \tab 156 \tab 112 \tab 8 \tab 1372 \cr
   Total \tab 0 \tab 251 \tab 572 \tab 273 \tab 156 \tab 112 \tab 8 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influhsp}},
\code{\link{influso}},
\code{\link{racwork}}
}
\concept{Intergroup Relations}
\keyword{variable}
