% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_371.R
\name{hsphys}
\alias{hsphys}
\title{R ever took a high school physics course}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsphys
}
\details{
Question Did you take a high school physics course?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3354/vshow}{GSS Data Explorer page for \code{hsphys}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2690 \tab 555 \tab 1222 \tab 11 \tab 32 \tab 4510 \cr
   2008 \tab 597 \tab 432 \tab 979 \tab 12 \tab 3 \tab 2023 \cr
   2010 \tab 1370 \tab 203 \tab 406 \tab 1 \tab 64 \tab 2044 \cr
   2012 \tab 1068 \tab 307 \tab 577 \tab 19 \tab 3 \tab 1974 \cr
   2014 \tab 1348 \tab 382 \tab 796 \tab 11 \tab 1 \tab 2538 \cr
   2016 \tab 1532 \tab 416 \tab 898 \tab 13 \tab 8 \tab 2867 \cr
   2018 \tab 1222 \tab 398 \tab 711 \tab 14 \tab 3 \tab 2348 \cr
   Total \tab 9827 \tab 2693 \tab 5589 \tab 81 \tab 114 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsmath}},
\code{\link{sciagrgw}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
