% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{hsrespct}
\alias{hsrespct}
\title{School subjects: respect for authority}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsrespct
}
\details{
Question Here are some things that might be taught in school. How important is it that schools teach each of these to 15-year-olds?
C. Respect for authority.
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4017/vshow}{GSS Data Explorer page for \code{hsrespct}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 316 \tab 215 \tab 99 \tab 17 \tab 9 \tab 9 \tab 12 \tab 1534 \cr
   Total \tab 857 \tab 316 \tab 215 \tab 99 \tab 17 \tab 9 \tab 9 \tab 12 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hssexed}},
\code{\link{opfamily}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
