% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{hssci}
\alias{hssci}
\title{School subjects: science and technology}
\source{
General Social Survey https://gss.norc.org
}
\description{
hssci
}
\details{
Question Here are some things that might be taught in school. How important is it that schools teach each of these to 15-year-olds?
G. Science and technology.
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4021/vshow}{GSS Data Explorer page for \code{hssci}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 249 \tab 251 \tab 135 \tab 19 \tab 11 \tab 12 \tab 1534 \cr
   Total \tab 857 \tab 249 \tab 251 \tab 135 \tab 19 \tab 11 \tab 12 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
