% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{huadd}
\alias{huadd}
\title{Housing unit (hu) at address?}
\source{
General Social Survey https://gss.norc.org
}
\description{
huadd
}
\details{
Question IS THERE A HOUSING UNIT (HU) AT THE ADDRESS SHOWN ON THE ASSIGNMENT LABEL?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6652/vshow}{GSS Data Explorer page for \code{huadd}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab - \tab 2838 \tab 26 \tab 2 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 2342 \tab 6 \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 2076 \tab 1467 \tab 1 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 2076 \tab 6647 \tab 33 \tab 2 \tab 1 \tab 0 \tab 8759 \cr
}
}

\keyword{variable}
