% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{hubbywk1}
\alias{hubbywk1}
\title{Men should earn money women keep house}
\source{
General Social Survey https://gss.norc.org
}
\description{
hubbywk1
}
\details{
Question And, do you agree or disagree . . .
A. A man''s job is to earn money; a woman''s job is to look after the home and family.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4402/vshow}{GSS Data Explorer page for \code{hubbywk1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1994 \tab 1545 \tab 76 \tab 227 \tab 271 \tab 544 \tab 296 \tab 8 \tab 25 \tab - \tab 2992 \cr
   2002 \tab 1594 \tab 101 \tab 172 \tab 276 \tab - \tab 608 \tab 13 \tab 1 \tab - \tab 2765 \cr
   2012 \tab 672 \tab 63 \tab 225 \tab 223 \tab 538 \tab 228 \tab 18 \tab 7 \tab - \tab 1974 \cr
   2022 \tab 2379 \tab 40 \tab 129 \tab 255 \tab 356 \tab 356 \tab 18 \tab 4 \tab 7 \tab 3544 \cr
   Total \tab 6190 \tab 280 \tab 753 \tab 1025 \tab 1438 \tab 1488 \tab 57 \tab 37 \tab 7 \tab 11275 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabfst}},
\code{\link{cohabok}},
\code{\link{divbest}},
\code{\link{divifkd1}},
\code{\link{divnokd1}},
\code{\link{fewrksup}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{singlpar}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
