% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{hubbywrk}
\alias{hubbywrk}
\title{Husb shld work wife shld look after home}
\source{
General Social Survey https://gss.norc.org
}
\description{
hubbywrk
}
\details{
Question Do you agree or disagree ...
I. A husband''s job is to earn money; a wife''s job is to look after the home and family.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1988, 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4351/vshow}{GSS Data Explorer page for \code{hubbywrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 119 \tab 266 \tab 281 \tab 441 \tab 265 \tab 20 \tab 22 \tab 1481 \cr
   1991 \tab 158 \tab 145 \tab 233 \tab 293 \tab 341 \tab 302 \tab 18 \tab 27 \tab 1517 \cr
   1998 \tab 1548 \tab 121 \tab 153 \tab 237 \tab 305 \tab 420 \tab 23 \tab 25 \tab 2832 \cr
   2008 \tab 658 \tab 159 \tab 216 \tab 283 \tab 413 \tab 287 \tab 6 \tab 1 \tab 2023 \cr
   2018 \tab 1173 \tab 73 \tab 157 \tab 213 \tab 351 \tab 370 \tab 5 \tab 6 \tab 2348 \cr
   Total \tab 3604 \tab 617 \tab 1025 \tab 1307 \tab 1851 \tab 1644 \tab 72 \tab 81 \tab 10201 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famsuffr}},
\code{\link{fejobind}},
\code{\link{hapifwrk}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{kidsuffr}},
\code{\link{mawrkwrm}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
