% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{hunt1}
\alias{hunt1}
\title{Does r or spouse or partner hunt}
\source{
General Social Survey https://gss.norc.org
}
\description{
hunt1
}
\details{
Question Do you (or does your (husband/wife/partner)) go hunting?
}
\section{Values}{

\itemize{
\item \code{1} yes, respondent does
\item \code{2} yes, spouse or partner does
\item \code{3} yes, both do
\item \code{4} no, neither r nor spouse or partner does
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/C/- \tab partial \cr
   2018 \tab A/C/- \tab partial \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/C/- \tab partial \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6195/vshow}{GSS Data Explorer page for \code{hunt1}}.
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}},
\code{\link{tvhours}}
}
\concept{Core}
\keyword{variable}
