% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{hurtoth}
\alias{hurtoth}
\title{How likely x violent toward other people}
\source{
General Social Survey https://gss.norc.org
}
\description{
hurtoth
}
\details{
Question In your opinion, how likely is it NAME would do something violent toward other people? Is it:
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2009/vshow}{GSS Data Explorer page for \code{hurtoth}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1460 \tab 231 \tab 487 \tab 427 \tab 187 \tab 91 \tab 21 \tab - \tab - \tab 2904 \cr
   2006 \tab 3073 \tab 143 \tab 469 \tab 554 \tab 180 \tab 76 \tab 15 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 133 \tab 415 \tab 419 \tab 130 \tab 46 \tab 30 \tab - \tab - \tab 2348 \cr
   2024 \tab 1692 \tab 119 \tab 539 \tab 695 \tab 157 \tab 64 \tab 13 \tab - \tab 30 \tab 3309 \cr
   Total \tab 7400 \tab 626 \tab 1910 \tab 2095 \tab 654 \tab 277 \tab 79 \tab 0 \tab 30 \tab 13071 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}

Other Mental Health: 
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
