% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{hwduties}
\alias{hwduties}
\title{Is r mainly respons.for hh chores?}
\source{
General Social Survey https://gss.norc.org
}
\description{
hwduties
}
\details{
Question Are you the person who is responsible for doing the general domestic duties - like cleaning, cooking, washing and so on - in your household?
}
\section{Values}{

\itemize{
\item \code{1} yes i responsible
\item \code{2} yes me and other
\item \code{3} no someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4495/vshow}{GSS Data Explorer page for \code{hwduties}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes i responsible \tab yes me and other \tab no someone else \tab no answer \tab don't know \tab iap \tab Total \cr
   1989 \tab - \tab 765 \tab 366 \tab 299 \tab 107 \tab - \tab - \tab 1537 \cr
   1998 \tab - \tab 687 \tab 301 \tab 185 \tab 213 \tab 1 \tab 1445 \tab 2832 \cr
   Total \tab 0 \tab 1452 \tab 667 \tab 484 \tab 320 \tab 1 \tab 1445 \tab 4369 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{paysame}},
\code{\link{sexemp}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}},
\code{\link{yrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
