% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{hyperlnk}
\alias{hyperlnk}
\title{How often r uses links in other sites}
\source{
General Social Survey https://gss.norc.org
}
\description{
hyperlnk
}
\details{
Question Now I'd like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit:
F. Go from a hyperlink in another site that you are visiting?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than half of the time
\item \verb{[3]} about half of the time
\item \verb{[4]} more than half of the time
\item \verb{[5]} almost always
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2393/vshow}{GSS Data Explorer page for \code{hyperlnk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than half of the time \tab about half of the time \tab more than half of the time \tab almost always \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 179 \tab 282 \tab 112 \tab 58 \tab 32 \tab 79 \tab - \tab 2817 \cr
   2002 \tab 2403 \tab 77 \tab 144 \tab 52 \tab 28 \tab 10 \tab 49 \tab 2 \tab 2765 \cr
   Total \tab 4478 \tab 256 \tab 426 \tab 164 \tab 86 \tab 42 \tab 128 \tab 2 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{bookmark}},
\code{\link{catdrcty}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{fin30}},
\code{\link{govt30}},
\code{\link{news30}},
\code{\link{schl30}},
\code{\link{work30}}
}
\concept{Information Society}
\keyword{variable}
