% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{hyperten}
\alias{hyperten}
\title{Told have hypertension or high blood pressure}
\source{
General Social Survey https://gss.norc.org
}
\description{
hyperten
}
\details{
Question Now I would like to ask you some questions about general health conditions. Has a doctor, nurse, or other health professional EVER told you that you had:
A. Hypertension or high blood pressure?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2842/vshow}{GSS Data Explorer page for \code{hyperten}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2014 \tab 1288 \tab 339 \tab 904 \tab 2 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 395 \tab 1016 \tab 2 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 1556 \tab 567 \tab 1392 \tab 8 \tab 6 \tab 15 \tab 3544 \cr
   Total \tab 3744 \tab 1301 \tab 3312 \tab 12 \tab 46 \tab 15 \tab 8430 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{knowschd}},
\code{\link{satjob1}},
\code{\link{slpprblm}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}
}
\concept{Quality of Working Life}
\keyword{variable}
