% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_379.R
\name{icesheet}
\alias{icesheet}
\title{True-false: the north pole is on a sheet of ice}
\source{
General Social Survey https://gss.norc.org
}
\description{
icesheet
}
\details{
Question The next few questions are about the Arctic and the Antarctic. The Arctic is the region around the North Pole; Antarctic is the region that contains the South Pole. These questions are like ones you might see on a television game show. If you don’t know or aren’t sure, just tell me so, and we will skip to the next question. Remember true, false, or don’t know.
A. The North Pole is on a sheet of ice that floats on the Arctic Ocean. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3432/vshow}{GSS Data Explorer page for \code{icesheet}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 756 \tab 426 \tab 681 \tab 1 \tab 4510 \cr
   2010 \tab 1338 \tab 328 \tab 175 \tab 191 \tab 12 \tab 2044 \cr
   Total \tab 3984 \tab 1084 \tab 601 \tab 872 \tab 13 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{comorsci}},
\code{\link{extinct}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{inuit}},
\code{\link{nosun}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
