% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{idols}
\alias{idols}
\title{Catholics are idols}
\source{
General Social Survey https://gss.norc.org
}
\description{
idols
}
\details{
Question To what extent to you agree or disagree with the following statements?
A. The statutes and images in Catholic churches are idols.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3094/vshow}{GSS Data Explorer page for \code{idols}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1475 \tab 168 \tab 324 \tab 448 \tab 235 \tab 115 \tab 30 \tab 17 \tab 2812 \cr
   Total \tab 1475 \tab 168 \tab 324 \tab 448 \tab 235 \tab 115 \tab 30 \tab 17 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{notthink}},
\code{\link{rosaries}}
}
\concept{Spirituality and Religion}
\keyword{variable}
