% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{idu30}
\alias{idu30}
\title{R inject drugs in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
idu30
}
\details{
Question How long has it been since you last used a needle to inject drugs?
}
\section{Values}{

\itemize{
\item \code{1} within the past 30 days
\item \code{2} more than 30 days ago but within the past 12 months
\item \code{3} more than 12 months ago but within the past 3 years
\item \code{4} more than 3 years ago
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5071/vshow}{GSS Data Explorer page for \code{idu30}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest2}},
\code{\link{relatsex}}

Other Drug Use And Addiction: 
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}}
}
\concept{Drug Use And Addiction}
\concept{High Risk Behaviors}
\keyword{variable}
