% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{if00who}
\alias{if00who}
\title{Who would r has voted for in 2000 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if00who
}
\details{
Question I. In 2000, you remember that Gore ran for President on the Democratic ticket against Bush for the Republicans . Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} gore
\item \code{2} bush
\item \code{3} nader
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/168/vshow}{GSS Data Explorer page for \code{if00who}}.
}

\seealso{
Other Core: 
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
