% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{if04who}
\alias{if04who}
\title{Who would r has voted for in 2004 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if04who
}
\details{
Question J. In 2004, you remember that Kerry ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election? 2. IF DID NOT VOTE OR INELIGIBLE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} kerry
\item \code{2} bush
\item \code{3} nader
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/171/vshow}{GSS Data Explorer page for \code{if04who}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab kerry \tab bush \tab nader \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3054 \tab 572 \tab 444 \tab 269 \tab 138 \tab 33 \tab 4510 \cr
   2008 \tab 1290 \tab 309 \tab 195 \tab 125 \tab 81 \tab 23 \tab 2023 \cr
   2010 \tab 1260 \tab 322 \tab 219 \tab 126 \tab 83 \tab 34 \tab 2044 \cr
   Total \tab 5604 \tab 1203 \tab 858 \tab 520 \tab 302 \tab 90 \tab 8577 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
