% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{if08who}
\alias{if08who}
\title{Who you would have voted for}
\source{
General Social Survey https://gss.norc.org
}
\description{
if08who
}
\details{
Question K. In 2008, you remember that Obama ran for President on the Democratic ticket against McCain for the Republicans. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:
Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} obama
\item \code{2} mccain
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/174/vshow}{GSS Data Explorer page for \code{if08who}}.
}

\seealso{
Other Core: 
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{pres12}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
