% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{if16who}
\alias{if16who}
\title{Who you would have voted for}
\source{
General Social Survey https://gss.norc.org
}
\description{
if16who
}
\details{
Question Who would you have voted for, for president, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} clinton
\item \code{2} trump
\item \code{3} other
\item \code{4} don't know/can't remember
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7091/vshow}{GSS Data Explorer page for \code{if16who}}.
}

\keyword{variable}
