% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{if20who}
\alias{if20who}
\title{Who you would have voted for in 2020}
\source{
General Social Survey https://gss.norc.org
}
\description{
if20who
}
\details{
Question Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} biden
\item \verb{[2]} trump
\item \verb{[3]} other
\item \verb{[4]} don't know/can't remember
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7760/vshow}{GSS Data Explorer page for \code{if20who}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab biden \tab trump \tab other \tab don't know/can't remember \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 260 \tab 315 \tab 135 \tab 102 \tab 24 \tab 2466 \tab 238 \tab 4 \tab 3544 \cr
   2024 \tab - \tab 231 \tab 316 \tab 125 \tab 120 \tab 23 \tab 2203 \tab 82 \tab 209 \tab 3309 \cr
   Total \tab 0 \tab 491 \tab 631 \tab 260 \tab 222 \tab 47 \tab 4669 \tab 320 \tab 213 \tab 6853 \cr
}
}

\keyword{variable}
