% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{if68who}
\alias{if68who}
\title{Who would r have voted for-1968 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if68who
}
\details{
Question A. Now in 1968, you remember that Humphrey ran for President on the Democratic ticket against Nixon for the Republicans, and Wallace as an Independent. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE: Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} humphrey
\item \code{2} nixon
\item \code{3} wallace
\item \code{4} other
\item \code{5} refused
\item \code{6} wldnt vt-relig
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
   1973 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/144/vshow}{GSS Data Explorer page for \code{if68who}}.
}

\seealso{
Other Core: 
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
