% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{if72who}
\alias{if72who}
\title{Who would r have voted for-1972 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if72who
}
\details{
Question B. In 1972, you remember that McGovern ran for President on the Democratic ticket against Nixon for the Republicans. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} mcgovern
\item \code{2} nixon
\item \code{3} other
\item \code{4} refused
\item \code{5} wouldnt vote
\item \code{6} wldnt vt-relig
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1975, 1976, 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/147/vshow}{GSS Data Explorer page for \code{if72who}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab mcgovern \tab nixon \tab other \tab wouldnt vote \tab don't know \tab no answer \tab refused \tab wldnt vt-relig \tab Total \cr
   1973 \tab 1053 \tab 120 \tab 223 \tab 40 \tab 2 \tab 44 \tab 22 \tab - \tab - \tab 1504 \cr
   1974 \tab 1032 \tab 160 \tab 145 \tab 66 \tab 16 \tab 58 \tab 6 \tab 1 \tab - \tab 1484 \cr
   1975 \tab 963 \tab 196 \tab 177 \tab 36 \tab 16 \tab 78 \tab 24 \tab - \tab - \tab 1490 \cr
   1976 \tab 975 \tab 189 \tab 169 \tab 61 \tab 5 \tab 92 \tab 7 \tab - \tab 1 \tab 1499 \cr
   1977 \tab 945 \tab 188 \tab 233 \tab 30 \tab 30 \tab 80 \tab 24 \tab - \tab - \tab 1530 \cr
   Total \tab 4968 \tab 853 \tab 947 \tab 233 \tab 69 \tab 352 \tab 83 \tab 1 \tab 1 \tab 7507 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
