% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{if80who}
\alias{if80who}
\title{Who would r have voted for-1980 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if80who
}
\details{
Question D. In 1980, you remember that Carter ran for President on the Democratic ticket against Reagan for the Republicans, and  anderson as an Independent.
Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} carter
\item \verb{[2]} reagan
\item \verb{[3]} anderson
\item \verb{[4]} other
\item \verb{[5]} wouldnt vote
\item \verb{[7]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984, 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/153/vshow}{GSS Data Explorer page for \code{if80who}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab carter \tab reagan \tab anderson \tab other \tab refused \tab don't know \tab no answer \tab wouldnt vote \tab Total \cr
   1982 \tab 1211 \tab 318 \tab 160 \tab 47 \tab 37 \tab 1 \tab 82 \tab 4 \tab - \tab 1860 \cr
   1983 \tab 1081 \tab 235 \tab 148 \tab 35 \tab 20 \tab - \tab 64 \tab 12 \tab 4 \tab 1599 \cr
   1984 \tab 967 \tab 219 \tab 161 \tab 15 \tab 20 \tab - \tab 71 \tab 20 \tab - \tab 1473 \cr
   1985 \tab 1040 \tab 165 \tab 233 \tab 27 \tab 9 \tab - \tab 48 \tab 8 \tab 4 \tab 1534 \cr
   1987 \tab 1216 \tab 260 \tab 182 \tab 35 \tab 6 \tab - \tab 87 \tab 33 \tab - \tab 1819 \cr
   Total \tab 5515 \tab 1197 \tab 884 \tab 159 \tab 92 \tab 1 \tab 352 \tab 77 \tab 8 \tab 8285 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
