% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{if84who}
\alias{if84who}
\title{Who would r have voted for-1984 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if84who
}
\details{
Question E. In 1984, you remember that Mondale ran for President on the Democratic ticket against Reagan for the Republicans. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:
Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} mondale
\item \code{2} reagan
\item \code{3} other
\item \code{4} wouldnt vote
\item \code{5} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989 \cr
   Full \tab 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/156/vshow}{GSS Data Explorer page for \code{if84who}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab mondale \tab reagan \tab other \tab wouldnt vote \tab don't know \tab no answer \tab Total \cr
   1985 \tab 1057 \tab 144 \tab 254 \tab 28 \tab 5 \tab 37 \tab 9 \tab 1534 \cr
   1986 \tab 1007 \tab 142 \tab 225 \tab 26 \tab 2 \tab 61 \tab 7 \tab 1470 \cr
   1987 \tab 1211 \tab 229 \tab 239 \tab 33 \tab - \tab 88 \tab 19 \tab 1819 \cr
   1988 \tab 942 \tab 156 \tab 250 \tab 38 \tab - \tab 88 \tab 7 \tab 1481 \cr
   1989 \tab 1025 \tab 111 \tab 234 \tab 28 \tab - \tab 62 \tab 77 \tab 1537 \cr
   Total \tab 5242 \tab 782 \tab 1202 \tab 153 \tab 7 \tab 336 \tab 119 \tab 7841 \cr
}
}

\seealso{
Other Core: 
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
