% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{if88who}
\alias{if88who}
\title{Who would r have voted for-1988 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if88who
}
\details{
Question F. In 1988, you remember that Dukakis ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:
Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} dukakis
\item \code{2} bush
\item \code{3} other
\item \code{4} wouldnt vote
\item \code{5} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/159/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab bush \tab don't know \tab dukakis \tab no answer \tab other \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1009 \tab 264 \tab 69 \tab 141 \tab 22 \tab 32 \tab - \tab 1537 \cr
   1990 \tab 913 \tab 252 \tab 72 \tab 104 \tab 9 \tab 22 \tab - \tab 1372 \cr
   1991 \tab 1017 \tab 292 \tab 74 \tab 86 \tab 20 \tab 28 \tab - \tab 1517 \cr
   1993 \tab 1045 \tab 282 \tab 112 \tab 93 \tab 40 \tab 34 \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 66798 \tab 1090 \tab 327 \tab 424 \tab 91 \tab 116 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}

Other Political: 
\code{\link{if84who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\concept{Political}
\keyword{variable}
