% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{if96who}
\alias{if96who}
\title{Who would r have voted for-1996 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if96who
}
\details{
Question H. In 1996, you remember that Clinton ran for President on the Democratic ticket against Dole for the Republicans and Perot for the Reform Party. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} clinton
\item \verb{[2]} dole
\item \verb{[3]} perot
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/165/vshow}{GSS Data Explorer page for \code{if96who}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab clinton \tab dole \tab perot \tab other \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1842 \tab 469 \tab 126 \tab 131 \tab 50 \tab 152 \tab 62 \tab 2832 \cr
   2000 \tab 1814 \tab 465 \tab 107 \tab 110 \tab 49 \tab 205 \tab 67 \tab 2817 \cr
   2002 \tab 1758 \tab 541 \tab 149 \tab 103 \tab 127 \tab 73 \tab 14 \tab 2765 \cr
   Total \tab 5414 \tab 1475 \tab 382 \tab 344 \tab 226 \tab 430 \tab 143 \tab 8414 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
