% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{ifwrong}
\alias{ifwrong}
\title{Agree people should support their country}
\source{
General Social Survey https://gss.norc.org
}
\description{
ifwrong
}
\details{
Question How much do you agree or disagree with the following statements?
E. People should support their country even if the country is in the wrong
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4838/vshow}{GSS Data Explorer page for \code{ifwrong}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1537 \tab 143 \tab 277 \tab 245 \tab 499 \tab 151 \tab 37 \tab 15 \tab - \tab - \tab 2904 \cr
   2004 \tab 1596 \tab 135 \tab 299 \tab 268 \tab 382 \tab 109 \tab 22 \tab 1 \tab - \tab - \tab 2812 \cr
   2014 \tab 1264 \tab 93 \tab 280 \tab 267 \tab 469 \tab 128 \tab 32 \tab 5 \tab - \tab - \tab 2538 \cr
   2024 \tab 1661 \tab 86 \tab 222 \tab 411 \tab 606 \tab 239 \tab 54 \tab 7 \tab - \tab 23 \tab 3309 \cr
   Total \tab 6058 \tab 457 \tab 1078 \tab 1191 \tab 1956 \tab 627 \tab 145 \tab 28 \tab 0 \tab 23 \tab 11563 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}

Other United States: 
\code{\link{amsports}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
