% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{ignorwk}
\alias{ignorwk}
\title{Feel ignored at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
ignorwk
}
\details{
Question I have felt ignored, excluded or isolated from others at work.
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3015/vshow}{GSS Data Explorer page for \code{ignorwk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2012 \tab 795 \tab 30 \tab 139 \tab 262 \tab 703 \tab 13 \tab 32 \tab 1974 \cr
   Total \tab 795 \tab 30 \tab 139 \tab 262 \tab 703 \tab 13 \tab 32 \tab 1974 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{denyrais}},
\code{\link{eharaswk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{wkbhvrs}},
\code{\link{wkrspns}}
}
\concept{Workplace Conflict}
\keyword{variable}
