% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{imbalnce}
\alias{imbalnce}
\title{How likely x s situation caused by imbalance in brain}
\source{
General Social Survey https://gss.norc.org
}
\description{
imbalnce
}
\details{
Question In your opinion, how likely is it that NAME's situation might be caused by (READ a-f) -- very likely, somewhat likely, not very likely, or not at all likely?
B. A chemical imbalance in the brain
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1989/vshow}{GSS Data Explorer page for \code{imbalnce}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 296 \tab 536 \tab 299 \tab 198 \tab 101 \tab 14 \tab 2904 \cr
   2006 \tab 3073 \tab 391 \tab 649 \tab 227 \tab 94 \tab 61 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 337 \tab 482 \tab 198 \tab 87 \tab 40 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 1024 \tab 1667 \tab 724 \tab 379 \tab 202 \tab 58 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
