% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{immcrime1}
\alias{immcrime1}
\title{Immigrants increase crime rates}
\source{
General Social Survey https://gss.norc.org
}
\description{
immcrime1
}
\details{
Question There are different opinions about immigrants from other countries living in America. By “immigrants” we mean people who come to settle in America. How much do you agree or disagree with each of the following statements?… Immigrants increase crime rates.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8062/vshow}{GSS Data Explorer page for \code{immcrime1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 127 \tab 296 \tab 401 \tab 496 \tab 225 \tab 79 \tab 1663 \tab 5 \tab 17 \tab 3309 \cr
   Total \tab 0 \tab 127 \tab 296 \tab 401 \tab 496 \tab 225 \tab 79 \tab 1663 \tab 5 \tab 17 \tab 3309 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{brneffrt}},
\code{\link{immideas1}},
\code{\link{immjobs1}}
}
\concept{ISSP National Identity}
\keyword{variable}
