% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{immeduc}
\alias{immeduc}
\title{Legal immigrants should have same education as americans}
\source{
General Social Survey https://gss.norc.org
}
\description{
immeduc
}
\details{
Question How much do you agree of disagree with the following statements?
C. Legal immigrants should have equal access to public education as American citizens.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4900/vshow}{GSS Data Explorer page for \code{immeduc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab 5 \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2014 \tab 1264 \tab 288 \tab 761 \tab 97 \tab 78 \tab 26 \tab 20 \tab 4 \tab - \tab 2538 \cr
   Total \tab 1264 \tab 288 \tab 761 \tab 97 \tab 78 \tab 26 \tab 20 \tab 4 \tab 0 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amproud1}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{belikeus}},
\code{\link{biasnews}},
\code{\link{bizmoney}},
\code{\link{brneffrt}},
\code{\link{citizen}},
\code{\link{citworld}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{decsorgs}},
\code{\link{ethclose}},
\code{\link{experts}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{freetrde}},
\code{\link{fulldem}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immassim}},
\code{\link{immcosts}},
\code{\link{immcrime1}},
\code{\link{immcult}},
\code{\link{immideas}},
\code{\link{immideas1}},
\code{\link{immimp}},
\code{\link{immjobs1}},
\code{\link{immrghts}},
\code{\link{immwlfare}},
\code{\link{imports}},
\code{\link{incgap1}},
\code{\link{internet}},
\code{\link{intlincs}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{lessprd}},
\code{\link{letin1}},
\code{\link{letin1a}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{refugees}},
\code{\link{res161}},
\code{\link{shortcom}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{wrldgovt}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{ISSP National Identity}
\concept{Immigration}
\keyword{variable}
