% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{immfate}
\alias{immfate}
\title{Fate of illegal immigrants}
\source{
General Social Survey https://gss.norc.org
}
\description{
immfate
}
\details{
Question What should be done about immigrants who are currently living in the U.S. illegally?
}
\section{Values}{

\itemize{
\item \verb{[1]} become citizens
\item \verb{[2]} become permanent legal residents
\item \verb{[3]} identified and deported
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7756/vshow}{GSS Data Explorer page for \code{immfate}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab become citizens \tab become permanent legal residents \tab identified and deported \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 1575 \tab 341 \tab 382 \tab 26 \tab 1203 \tab 7 \tab 10 \tab 3544 \cr
   2024 \tab - \tab 1303 \tab 320 \tab 547 \tab 45 \tab 1067 \tab 12 \tab 15 \tab 3309 \cr
   Total \tab 0 \tab 2878 \tab 661 \tab 929 \tab 71 \tab 2270 \tab 19 \tab 25 \tab 6853 \cr
}
}

\keyword{variable}
