% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{immideas1}
\alias{immideas1}
\title{Immigrants bring new ideas and cultures}
\source{
General Social Survey https://gss.norc.org
}
\description{
immideas1
}
\details{
Question How much do you agree or disagree with each of the following statements?...
Immigrants improve American society by bringing new ideas and cultures.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8063/vshow}{GSS Data Explorer page for \code{immideas1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 246 \tab 735 \tab 378 \tab 147 \tab 44 \tab 76 \tab 1663 \tab 2 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 246 \tab 735 \tab 378 \tab 147 \tab 44 \tab 76 \tab 1663 \tab 2 \tab 18 \tab 3309 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{brneffrt}},
\code{\link{immcrime1}},
\code{\link{immjobs1}}
}
\concept{ISSP National Identity}
\keyword{variable}
