% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{immimp}
\alias{immimp}
\title{Immigrants improve american society}
\source{
General Social Survey https://gss.norc.org
}
\description{
immimp
}
\details{
Question There are different opinions about immigrants from other countries living in America. (By "immigrants" we mean people who come to settle in America) How much do you agree or disagree with each of the following statements? (Please, select one response)
A. Immigrants improve American society by bringing in new ideas and cultures
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4890/vshow}{GSS Data Explorer page for \code{immimp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 119 \tab 563 \tab 294 \tab 176 \tab 40 \tab 23 \tab 1 \tab 2812 \cr
   Total \tab 1596 \tab 119 \tab 563 \tab 294 \tab 176 \tab 40 \tab 23 \tab 1 \tab 2812 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta2a}},
\code{\link{nafta3}}
}
\concept{ISSP National Identity}
\keyword{variable}
