% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{immstats}
\alias{immstats}
\title{R's current immigration status right now}
\source{
General Social Survey https://gss.norc.org
}
\description{
immstats
}
\details{
Question Thinking of your current immigration status right now, is that in Box A, B, or C?  If it’s in Box B, we do not want to know which specific category applies to you, and there will be no more questions about Box B. (IF RESPONDENT ASKS WHY, SAY: Because right now we are focusing on Box A.)
}
\section{Values}{

\itemize{
\item \code{1} box a
\item \code{2} box b
\item \code{3} box c
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3027/vshow}{GSS Data Explorer page for \code{immstats}}.
}

\seealso{
Other Immigration: 
\code{\link{bornhome}},
\code{\link{natborn}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}}

Other Immigration: 
\code{\link{bornhome}},
\code{\link{natborn}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}}
}
\concept{Immigration}
\keyword{variable}
