% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_173.R
\name{immunemp}
\alias{immunemp}
\title{Will immigrants fuel unemployment}
\source{
General Social Survey https://gss.norc.org
}
\description{
immunemp
}
\details{
Question What do you think will happen as a result of more immigrants coming to this country?  Is each of these possible results very likely, somewhat likely, not too likely, or not at all likely?
B. Higher unemployment.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not so likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1614/vshow}{GSS Data Explorer page for \code{immunemp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not so likely \tab not at all likely \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1518 \tab 795 \tab 458 \tab 116 \tab 34 \tab 63 \tab 8 \tab - \tab 2992 \cr
   Total \tab 1518 \tab 795 \tab 458 \tab 116 \tab 34 \tab 63 \tab 8 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{asngovt}},
\code{\link{bilinged}},
\code{\link{blkgovt}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{econfutr}},
\code{\link{econpast}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hspasn10}},
\code{\link{hspgovt}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{jobaff}},
\code{\link{letin}},
\code{\link{meltpot}},
\code{\link{obrespct}},
\code{\link{owneth}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{united}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Immigration}
\concept{Multiculturalism}
\keyword{variable}
