% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_173.R
\name{immunite}
\alias{immunite}
\title{Will immigrants affect national unity}
\source{
General Social Survey https://gss.norc.org
}
\description{
immunite
}
\details{
Question What do you think will happen as a result of more immigrants coming to this country?  Is each of these possible results very likely, somewhat likely, not too likely, or not at all likely?
C. Making it harder to keep the country united.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not so likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1615/vshow}{GSS Data Explorer page for \code{immunite}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not so likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 500 \tab 497 \tab 265 \tab 111 \tab 93 \tab 8 \tab 2992 \cr
   2000 \tab 1419 \tab 268 \tab 473 \tab 454 \tab 134 \tab 60 \tab 9 \tab 2817 \cr
   Total \tab 2937 \tab 768 \tab 970 \tab 719 \tab 245 \tab 153 \tab 17 \tab 5809 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{hspasn10}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immwrkup}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}}
}
\concept{Multiculturalism}
\keyword{variable}
