% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_153.R
\name{impgod}
\alias{impgod}
\title{Imp to r of having faith in god}
\source{
General Social Survey https://gss.norc.org
}
\description{
impgod
}
\details{
Question I'm going to read you a list of some things that different people value. Some people say these things are very important to them. Other people say they are not so important. Please tell me how important each things is to you personally, using the responses on this card (HAND CARD TO RESPONDENT). How about being financially secure? Is it one of the most important values you hold, very important, somewhat important, not too important, or not at all important?
D. Having faith in God.
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most important
\item \verb{[2]} very important
\item \verb{[3]} somewhat important
\item \verb{[4]} not too important
\item \verb{[5]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1441/vshow}{GSS Data Explorer page for \code{impgod}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab one of most important \tab very important \tab somewhat important \tab not too important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 733 \tab 506 \tab 189 \tab 90 \tab 68 \tab 6 \tab 14 \tab 1606 \cr
   Total \tab 0 \tab 733 \tab 506 \tab 189 \tab 90 \tab 68 \tab 6 \tab 14 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{impcultr}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{mostimp1}},
\code{\link{mostimp2}},
\code{\link{mostimp3}}
}
\concept{Culture}
\keyword{variable}
