% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{impgrn}
\alias{impgrn}
\title{More important in life than save environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
impgrn
}
\details{
Question How much do you agree or disagree with each of these statements?
A. There are more important things to do in life than protect the environment.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4779/vshow}{GSS Data Explorer page for \code{impgrn}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1541 \tab 33 \tab 204 \tab 346 \tab 479 \tab 123 \tab 25 \tab 66 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 38 \tab 301 \tab 299 \tab 616 \tab 134 \tab 32 \tab 10 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 54 \tab 262 \tab 553 \tab 664 \tab 255 \tab 57 \tab 179 \tab 5 \tab 4032 \cr
   Total \tab 4158 \tab 125 \tab 767 \tab 1198 \tab 1759 \tab 512 \tab 114 \tab 255 \tab 5 \tab 8893 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
