% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{implives}
\alias{implives}
\title{Way of life more important than being catholic}
\source{
General Social Survey https://gss.norc.org
}
\description{
implives
}
\details{
Question Please tell me whether you strongly agree, agree, disagree, or strongly disagree with the following statements:
B. How a person lives is more important than whether or not he or she is Catholic.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1159/vshow}{GSS Data Explorer page for \code{implives}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2135 \tab 238 \tab 326 \tab 62 \tab 10 \tab 22 \tab 24 \tab 2817 \cr
   Total \tab 2135 \tab 238 \tab 326 \tab 62 \tab 10 \tab 22 \tab 24 \tab 2817 \cr
}
}

\seealso{
Other Religion: 
\code{\link{chrissch}},
\code{\link{denomsch}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}}
}
\concept{Religion}
\keyword{variable}
