% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{imports}
\alias{imports}
\title{America should limit the import}
\source{
General Social Survey https://gss.norc.org
}
\description{
imports
}
\details{
Question How much do you agree or disagree with the following statements?
A. America should limit the import of foreign products in order to protect its national economy
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2021 \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4850/vshow}{GSS Data Explorer page for \code{imports}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1996 \tab 1537 \tab 291 \tab 589 \tab 219 \tab 142 \tab 40 \tab 62 \tab 24 \tab - \tab 2904 \cr
   2004 \tab 1596 \tab 271 \tab 453 \tab 253 \tab 180 \tab 23 \tab 35 \tab 1 \tab - \tab 2812 \cr
   2014 \tab 1264 \tab 231 \tab 490 \tab 258 \tab 209 \tab 32 \tab 50 \tab 4 \tab - \tab 2538 \cr
   2021 \tab 2003 \tab 195 \tab 617 \tab 607 \tab 326 \tab 55 \tab 66 \tab 149 \tab 14 \tab 4032 \cr
   Total \tab 6400 \tab 988 \tab 2149 \tab 1337 \tab 857 \tab 150 \tab 213 \tab 178 \tab 14 \tab 12286 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
