% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_85.R
\name{imprelax}
\alias{imprelax}
\title{Free time and relaxation}
\source{
General Social Survey https://gss.norc.org
}
\description{
imprelax
}
\details{
Question On these cards are various aspects of life. We would like to know how important each of these aspects of life is for you. On each of these Cards you see on the right hand side a scale with seven points. The lowest point with number 1 indicates that this aspect of life is unimportant to you. Point 7 at the top indicates that particular aspect of life is very important to you. The numbers in between indicate varying degrees of importance. Please take a look at all the cards first. Then, tell me for each card its letter and the number you've decided on. SHUFFLE THE SET OF BLUE CardS BEFORE HANDING TO RESPONDENT.
C. Free time and relaxation.
}
\section{Values}{

\itemize{
\item \code{1} unimportant
\item \code{7} very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/767/vshow}{GSS Data Explorer page for \code{imprelax}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab unimportant \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 34 \tab 58 \tab 98 \tab 188 \tab 429 \tab 349 \tab 678 \tab 6 \tab 20 \tab 1860 \cr
   Total \tab 0 \tab 34 \tab 58 \tab 98 \tab 188 \tab 429 \tab 349 \tab 678 \tab 6 \tab 20 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{impwork}}
}
\concept{Core}
\keyword{variable}
