% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_328.R
\name{imprvmed}
\alias{imprvmed}
\title{Situation will improve with medical treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
imprvmed
}
\details{
Question L. In your opinion, how likely is it that this youth's/child's situation will improve with medical treatment?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2932/vshow}{GSS Data Explorer page for \code{imprvmed}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 516 \tab 460 \tab 226 \tab 137 \tab 42 \tab 12 \tab 2765 \cr
   Total \tab 1372 \tab 516 \tab 460 \tab 226 \tab 137 \tab 42 \tab 12 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{allergic}},
\code{\link{bettrown}},
\code{\link{discipln}},
\code{\link{geneprob}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{nextdoor}},
\code{\link{raised}},
\code{\link{violtv}}
}
\concept{Mental Health II}
\keyword{variable}
