% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{imprvown}
\alias{imprvown}
\title{How likely x s situation will improve on its own}
\source{
General Social Survey https://gss.norc.org
}
\description{
imprvown
}
\details{
Question In your opinion, how likely is it that NAME's situation will improve on its own -- very likely, somewhat likely, somewhat unlikely, or not likely at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2001/vshow}{GSS Data Explorer page for \code{imprvown}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 137 \tab 256 \tab 335 \tab 668 \tab 34 \tab 14 \tab 2904 \cr
   2006 \tab 3073 \tab 113 \tab 324 \tab 529 \tab 407 \tab 49 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 89 \tab 247 \tab 392 \tab 395 \tab 22 \tab 28 \tab 2348 \cr
   Total \tab 5708 \tab 339 \tab 827 \tab 1256 \tab 1470 \tab 105 \tab 57 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
