% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{incentiv}
\alias{incentiv}
\title{Pay differences needed for motivation}
\source{
General Social Survey https://gss.norc.org
}
\description{
incentiv
}
\details{
Question Some people earn a lot of money while others do not earn very much at all. In order to get people to work hard, do you think large differences in pay are . . .
}
\section{Values}{

\itemize{
\item \code{1} absolutely neccessary
\item \code{2} probably neccessary
\item \code{3} probably not necc
\item \code{4} definitely not necc
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4255/vshow}{GSS Data Explorer page for \code{incentiv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab absolutely neccessary \tab probably neccessary \tab probably not necc \tab definitely not necc \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 302 \tab 759 \tab 301 \tab 97 \tab 80 \tab 25 \tab 1819 \cr
   Total \tab 255 \tab 302 \tab 759 \tab 301 \tab 97 \tab 80 \tab 25 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{oppol}},
\code{\link{opsex}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
