% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_472.R
\name{incgap}
\alias{incgap}
\title{Income differentials in usa too big}
\source{
General Social Survey https://gss.norc.org
}
\description{
incgap
}
\details{
Question Do you agree or disagree?
A. Differences in income in America are too large.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2000 \cr
   B/C/- \tab 2008, 2018, 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4295/vshow}{GSS Data Explorer page for \code{incgap}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 244 \tab 636 \tab 324 \tab 240 \tab 46 \tab 50 \tab 24 \tab - \tab 1819 \cr
   1996 \tab 1460 \tab 464 \tab 467 \tab 178 \tab 169 \tab 117 \tab 41 \tab 8 \tab - \tab 2904 \cr
   2000 \tab 1545 \tab 297 \tab 489 \tab 255 \tab 109 \tab 38 \tab 37 \tab 47 \tab - \tab 2817 \cr
   2008 \tab 982 \tab 259 \tab 395 \tab 204 \tab 135 \tab 33 \tab 10 \tab 5 \tab - \tab 2023 \cr
   2018 \tab 1175 \tab 349 \tab 379 \tab 266 \tab 118 \tab 31 \tab 29 \tab 1 \tab - \tab 2348 \cr
   2021 \tab 2029 \tab 793 \tab 545 \tab 301 \tab 130 \tab 53 \tab 41 \tab 134 \tab 6 \tab 4032 \cr
   Total \tab 7446 \tab 2406 \tab 2911 \tab 1528 \tab 901 \tab 318 \tab 208 \tab 219 \tab 6 \tab 15943 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givrocc}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}}
}
\concept{ISSP Social Inequality}
\concept{Markets}
\keyword{variable}
