% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{incgap1}
\alias{incgap1}
\title{Government's responsibility to reduce economic inequality}
\source{
General Social Survey https://gss.norc.org
}
\description{
incgap1
}
\details{
Question Thinking about the situation in America. To what extent do you agree or disagree with the following statement? It should be the government’s responsibility to reduce income differences between the rich and the poor.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8051/vshow}{GSS Data Explorer page for \code{incgap1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 247 \tab 415 \tab 351 \tab 279 \tab 233 \tab 95 \tab 1664 \tab 6 \tab 19 \tab 3309 \cr
   Total \tab 0 \tab 247 \tab 415 \tab 351 \tab 279 \tab 233 \tab 95 \tab 1664 \tab 6 \tab 19 \tab 3309 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{bizmoney}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{experts}},
\code{\link{fulldem}},
\code{\link{immwlfare}}
}
\concept{ISSP National Identity}
\keyword{variable}
