% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{inchrw}
\alias{inchrw}
\title{Hours per week using work computer make money}
\source{
General Social Survey https://gss.norc.org
}
\description{
inchrw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
B. Other income-producing activities?
2. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2570/vshow}{GSS Data Explorer page for \code{inchrw}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 2 \tab 20 \tab 30 \tab 5 \tab 6 \tab 8 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2486 \tab 233 \tab 4 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 80 \tab 2817 \cr
   Total \tab 2486 \tab 233 \tab 4 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 80 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{ecomminw}},
\code{\link{emhrw}},
\code{\link{emminw}},
\code{\link{incminw}},
\code{\link{perhrw}},
\code{\link{perminw}},
\code{\link{wwwhrw}},
\code{\link{wwwminw}}
}
\concept{Information Society}
\keyword{variable}
