% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{income16}
\alias{income16}
\title{Total family income}
\source{
General Social Survey https://gss.norc.org
}
\description{
income16
}
\details{
Question In which of these groups did your total family income, from all sources, fall last year -- 2015 -- before taxes, that is. Just tell me the letter.                       
Total income includes interest or dividends, rent, Social Security, other pensions,  alimony or child support, unemployment compensation, public aid (welfare),  armed forces or veteran's allotment.
}
\section{Values}{

\itemize{
\item \verb{[1]} under $1,000
\item \verb{[2]} $1,000 to $2,999
\item \verb{[3]} $3,000 to $3,999
\item \verb{[4]} $4,000 to $4,999
\item \verb{[5]} $5,000 to $5,999
\item \verb{[6]} $6,000 to $6,999
\item \verb{[7]} $7,000 to $7,999
\item \verb{[8]} $8,000 to $9,999
\item \verb{[9]} $10,000 to $12,499
\item \verb{[10]} $12,500 to $14,999
\item \verb{[11]} $15,000 to $17,499
\item \verb{[12]} $17,500 to $19,999
\item \verb{[13]} $20,000 to $22,499
\item \verb{[14]} $22,500 to $24,999
\item \verb{[15]} $25,000 to $29,999
\item \verb{[16]} $30,000 to $34,999
\item \verb{[17]} $35,000 to $39,999
\item \verb{[18]} $40,000 to $49,999
\item \verb{[19]} $50,000 to $59,999
\item \verb{[20]} $60,000 to $74,999
\item \verb{[21]} $75,000 to $89,999
\item \verb{[22]} $90,000 to $109,999
\item \verb{[23]} $110,000 to $129,999
\item \verb{[24]} $130,000 to $149,999
\item \verb{[25]} $150,000 to $169,999
\item \verb{[26]} $170,000 or over
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6167/vshow}{GSS Data Explorer page for \code{income16}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab under $1,000 \tab $1,000 to $2,999 \tab $3,000 to $3,999 \tab $4,000 to $4,999 \tab $5,000 to $5,999 \tab $6,000 to $6,999 \tab $7,000 to $7,999 \tab $8,000 to $9,999 \tab $10,000 to $12,499 \tab $12,500 to $14,999 \tab $15,000 to $17,499 \tab $17,500 to $19,999 \tab $20,000 to $22,499 \tab $22,500 to $24,999 \tab $25,000 to $29,999 \tab $30,000 to $34,999 \tab $35,000 to $39,999 \tab $40,000 to $49,999 \tab $50,000 to $59,999 \tab $60,000 to $74,999 \tab $75,000 to $89,999 \tab $90,000 to $109,999 \tab $110,000 to $129,999 \tab $130,000 to $149,999 \tab $150,000 to $169,999 \tab $170,000 or over \tab don't know \tab refused \tab skipped on web \tab Total \cr
   2016 \tab - \tab 36 \tab 35 \tab 24 \tab 13 \tab 19 \tab 15 \tab 23 \tab 48 \tab 88 \tab 71 \tab 65 \tab 65 \tab 95 \tab 101 \tab 115 \tab 131 \tab 126 \tab 218 \tab 207 \tab 258 \tab 214 \tab 179 \tab 126 \tab 99 \tab 61 \tab 164 \tab 134 \tab 137 \tab - \tab 2867 \cr
   2018 \tab - \tab 33 \tab 27 \tab 19 \tab 14 \tab 15 \tab 9 \tab 13 \tab 41 \tab 70 \tab 70 \tab 44 \tab 37 \tab 65 \tab 77 \tab 80 \tab 116 \tab 101 \tab 159 \tab 187 \tab 215 \tab 167 \tab 145 \tab 128 \tab 77 \tab 72 \tab 171 \tab 91 \tab 105 \tab - \tab 2348 \cr
   2021 \tab - \tab 64 \tab 50 \tab 19 \tab 11 \tab 17 \tab 12 \tab 18 \tab 54 \tab 87 \tab 80 \tab 74 \tab 66 \tab 80 \tab 69 \tab 131 \tab 155 \tab 136 \tab 266 \tab 253 \tab 367 \tab 314 \tab 326 \tab 207 \tab 141 \tab 140 \tab 372 \tab 335 \tab 80 \tab 108 \tab 4032 \cr
   2022 \tab - \tab 39 \tab 33 \tab 18 \tab 11 \tab 16 \tab 14 \tab 16 \tab 40 \tab 104 \tab 83 \tab 64 \tab 51 \tab 92 \tab 100 \tab 129 \tab 162 \tab 152 \tab 211 \tab 239 \tab 304 \tab 240 \tab 249 \tab 169 \tab 147 \tab 97 \tab 330 \tab 268 \tab 132 \tab 34 \tab 3544 \cr
   Total \tab 0 \tab 172 \tab 145 \tab 80 \tab 49 \tab 67 \tab 50 \tab 70 \tab 183 \tab 349 \tab 304 \tab 247 \tab 219 \tab 332 \tab 347 \tab 455 \tab 564 \tab 515 \tab 854 \tab 886 \tab 1144 \tab 935 \tab 899 \tab 630 \tab 464 \tab 370 \tab 1037 \tab 828 \tab 454 \tab 142 \tab 12791 \cr
}
}

\seealso{
Other Core: 
\code{\link{dotdata}},
\code{\link{income06}},
\code{\link{region}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom98}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Core}
\keyword{variable}
