% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_16.R
\name{income72}
\alias{income72}
\title{Total family income}
\source{
General Social Survey https://gss.norc.org
}
\description{
income72
}
\details{
Question In which of these groups did your total family income, from all sources, fall last year--1971--before taxes, that is?
}
\section{Values}{

\itemize{
\item \verb{[1]} lt $2000
\item \verb{[2]} $2000 to $3999
\item \verb{[3]} $4000 to $5999
\item \verb{[4]} $6000 to $7999
\item \verb{[5]} $8000 to $9999
\item \verb{[6]} $10000 to $12499
\item \verb{[7]} $12500 to $14999
\item \verb{[8]} $15000 to $17499
\item \verb{[9]} $17500 to $19999
\item \verb{[10]} $20000 to $24999
\item \verb{[11]} $25000 to $29999
\item \verb{[12]} $30000+
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/106/vshow}{GSS Data Explorer page for \code{income72}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab lt $2000 \tab $2000 to $3999 \tab $4000 to $5999 \tab $6000 to $7999 \tab $8000 to $9999 \tab $10000 to $12499 \tab $12500 to $14999 \tab $15000 to $17499 \tab $17500 to $19999 \tab $20000 to $24999 \tab $25000 to $29999 \tab $30000+ \tab don't know \tab no answer \tab refused \tab iap \tab Total \cr
   1972 \tab 123 \tab 165 \tab 160 \tab 160 \tab 192 \tab 217 \tab 159 \tab 103 \tab 63 \tab 67 \tab 33 \tab 32 \tab 64 \tab 13 \tab 62 \tab - \tab 1613 \cr
   Total \tab 123 \tab 165 \tab 160 \tab 160 \tab 192 \tab 217 \tab 159 \tab 103 \tab 63 \tab 67 \tab 33 \tab 32 \tab 64 \tab 13 \tab 62 \tab 0 \tab 1613 \cr
}
}

\seealso{
Other Core: 
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}}
}
\concept{Core}
\keyword{variable}
