% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{incuspop}
\alias{incuspop}
\title{Estimated income status of housing unit}
\source{
General Social Survey https://gss.norc.org
}
\description{
incuspop
}
\details{
Question What do you believe the income status of this housing unit is, relative to the rest of the general United States population?
}
\section{Values}{

\itemize{
\item \verb{[1]} higher than average
\item \verb{[2]} average
\item \verb{[3]} lower than average
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6657/vshow}{GSS Data Explorer page for \code{incuspop}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab higher than average \tab average \tab lower than average \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab - \tab 518 \tab 1618 \tab 702 \tab 2 \tab 27 \tab - \tab 2867 \cr
   2018 \tab - \tab 385 \tab 1348 \tab 607 \tab 2 \tab 6 \tab - \tab 2348 \cr
   2022 \tab 2077 \tab 269 \tab 857 \tab 341 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 2077 \tab 1172 \tab 3823 \tab 1650 \tab 4 \tab 33 \tab 0 \tab 8759 \cr
}
}

\keyword{variable}
