% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_321.R
\name{indperf}
\alias{indperf}
\title{Size of perf based pay depend on individual}
\source{
General Social Survey https://gss.norc.org
}
\description{
indperf
}
\details{
Question C. Does the size of these performance-based payments depend on individual performance?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2014, 2018 \cr
   B/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2858/vshow}{GSS Data Explorer page for \code{indperf}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 2152 \tab 380 \tab 206 \tab 11 \tab 16 \tab - \tab 2765 \cr
   2006 \tab 3849 \tab 462 \tab 177 \tab 5 \tab 17 \tab - \tab 4510 \cr
   2014 \tab 2275 \tab 195 \tab 63 \tab 3 \tab 2 \tab - \tab 2538 \cr
   2018 \tab 1846 \tab 382 \tab 114 \tab 4 \tab 2 \tab - \tab 2348 \cr
   2022 \tab 3222 \tab 236 \tab 68 \tab 17 \tab - \tab 1 \tab 3544 \cr
   Total \tab 13344 \tab 1655 \tab 628 \tab 40 \tab 37 \tab 1 \tab 15705 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{deptperf}},
\code{\link{extr2001}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{seecowrk}},
\code{\link{yearval}}
}
\concept{Employee Compensation}
\keyword{variable}
