% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{indtag}
\alias{indtag}
\title{Indus10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
indtag
}
\details{
Question INDUS10 Coding Tag
}
\section{Values}{

\itemize{
\item \verb{[1]} found
\item \verb{[2]} not found, assigned
\item \verb{[3]} not found, not assigned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5985/vshow}{GSS Data Explorer page for \code{indtag}}.

Counts by year:\tabular{llllll}{
   year \tab found \tab not found, assigned \tab not found, not assigned \tab iap \tab Total \cr
   1972 \tab 702 \tab 751 \tab 160 \tab - \tab 1613 \cr
   1973 \tab 1234 \tab 197 \tab 73 \tab - \tab 1504 \cr
   1974 \tab 1217 \tab 230 \tab 37 \tab - \tab 1484 \cr
   1975 \tab 976 \tab 425 \tab 89 \tab - \tab 1490 \cr
   1976 \tab 1230 \tab 215 \tab 54 \tab - \tab 1499 \cr
   1977 \tab 817 \tab 593 \tab 120 \tab - \tab 1530 \cr
   1978 \tab 1518 \tab 14 \tab - \tab - \tab 1532 \cr
   1980 \tab 1293 \tab 146 \tab 29 \tab - \tab 1468 \cr
   1982 \tab 1696 \tab 145 \tab 19 \tab - \tab 1860 \cr
   1983 \tab 1475 \tab 106 \tab 18 \tab - \tab 1599 \cr
   1984 \tab 1458 \tab 14 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 1296 \tab 191 \tab 47 \tab - \tab 1534 \cr
   1986 \tab 1425 \tab 39 \tab 6 \tab - \tab 1470 \cr
   1987 \tab 1511 \tab 252 \tab 56 \tab - \tab 1819 \cr
   1988 \tab 1460 \tab 20 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 1420 \tab 106 \tab 11 \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 14 \tab - \tab - \tab 1372 \cr
   1991 \tab 1499 \tab 18 \tab - \tab - \tab 1517 \cr
   1993 \tab 1545 \tab 56 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 2938 \tab 47 \tab 7 \tab - \tab 2992 \cr
   1996 \tab 2883 \tab 20 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 2801 \tab 29 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   Total \tab 50723 \tab 3628 \tab 736 \tab 0 \tab 55087 \cr
}
}

\seealso{
Other Core: 
\code{\link{agewed}},
\code{\link{indstatus}},
\code{\link{indus10}},
\code{\link{marital}},
\code{\link{martype}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{prestg10}},
\code{\link{prestg105plus}}
}
\concept{Core}
\keyword{variable}
