% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_02.R
\name{indus07}
\alias{indus07}
\title{Industry code based on naics 2007}
\source{
General Social Survey https://gss.norc.org
}
\description{
indus07
}
\details{
Question C. What kind of place (do/did) you work for?
D. What (do/did) they (make/do)?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010 \cr
}
}

\seealso{
Other Core: 
\code{\link{commute}},
\code{\link{found}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{prestg80}}
}
\concept{Core}
\keyword{variable}
