% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_02.R
\name{indus80}
\alias{indus80}
\title{R's industry code (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
indus80
}
\details{
Question Respondent's Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item \verb{[10]} agricultural production, crops
\item \verb{[11]} agricultural production, livestock
\item \verb{[20]} agricultural services, except horticultural
\item \verb{[21]} horticultural services
\item \verb{[30]} forestry
\item \verb{[31]} fishing, hunting, and trapping, etc.
\item \verb{[40]} metal mining
\item \verb{[41]} coal mining
\item \verb{[42]} crude petroleum and natural gas extraction
\item \verb{[50]} nonmetallic mining and quarrying, except fuel
\item \verb{[60]} construction
\item \verb{[100]} meat products
\item \verb{[101]} dairy products
\item \verb{[102]} canned and preserved fruits and vegetables
\item \verb{[110]} grain mill products
\item \verb{[111]} bakery products
\item \verb{[112]} sugar and confectionery products
\item \verb{[120]} beverage industries
\item \verb{[121]} miscellaneous food preparations and kindred products
\item \verb{[122]} not specified food industries
\item \verb{[130]} tobacco manufactures
\item \verb{[132]} knitting mills
\item \verb{[140]} dyeing and finishing textiles, except wool and knit goods
\item \verb{[141]} floor coverings, except hard surface
\item \verb{[142]} yarn, thread, and fabric mills
\item \verb{[150]} miscellaneous textile mill products
\item \verb{[151]} apparel and accessories, except knit
\item \verb{[152]} miscellaneous fabricated textile products
\item \verb{[160]} pulp, paper, and paperboard mills
\item \verb{[161]} miscellaneous paper and pulp products
\item \verb{[162]} paperboard containers and boxes
\item \verb{[171]} newspaper publishing and printing
\item \verb{[172]} printing, publishing, and allied industries, except newspapers
\item \verb{[180]} plastics, synthetics, and resins
\item \verb{[181]} drugs
\item \verb{[182]} soaps and cosmetics
\item \verb{[185]} paints, varnishes, and related products
\item \verb{[190]} agricultural chemicals
\item \verb{[191]} industrial and miscellaneous chemicals
\item \verb{[192]} petroleum and coal products
\item \verb{[200]} petroleum refining
\item \verb{[201]} miscellaneous petroleum and coal products
\item \verb{[210]} tires and inner tubes
\item \verb{[211]} other rubber products, and plastics footwear and belting
\item \verb{[212]} miscellaneous plastics products
\item \verb{[220]} leather tanning and finishing
\item \verb{[221]} footwear, except rubber and plastic
\item \verb{[222]} leather products, except footwear
\item \verb{[230]} logging
\item \verb{[231]} sawmills, planning mills, and millwork
\item \verb{[232]} wood buildings and mobile homes
\item \verb{[241]} miscellaneous wood products
\item \verb{[242]} furniture and fixtures
\item \verb{[250]} glass and glass products
\item \verb{[251]} cement, concrete, gypsum, and plaster products
\item \verb{[252]} structural clay products
\item \verb{[261]} pottery and related products
\item \verb{[262]} miscellaneous nonmetallic mineral and stone products
\item \verb{[270]} blast furnaces, steelworks, rolling and finishing mills
\item \verb{[271]} iron and steel foundries
\item \verb{[272]} primary aluminum industries
\item \verb{[280]} other primary metal industries
\item \verb{[281]} cutlery, hand tools, and other hardware
\item \verb{[282]} fabricated structural metal products
\item \verb{[290]} screw machine products
\item \verb{[291]} metal forgings and stampings
\item \verb{[292]} ordnance
\item \verb{[300]} miscellaneous fabricated metal products
\item \verb{[301]} not specified metal industries
\item \verb{[310]} engines and turbines
\item \verb{[311]} farm machinery and equipment
\item \verb{[312]} construction and material handling machines
\item \verb{[320]} metalworking machinery
\item \verb{[321]} office and accounting machines
\item \verb{[322]} electronic computing equipment
\item \verb{[331]} machinery, except electrical, n.e.c.
\item \verb{[332]} not specified machinery
\item \verb{[340]} household appliances
\item \verb{[341]} radio, tv, and communication equipment
\item \verb{[342]} electrical machinery, equipment, and supplies, n.e.c.
\item \verb{[350]} not specified electrical machinery, equipment, and supplies
\item \verb{[351]} motor vehicles and motor vehicle equipment
\item \verb{[352]} aircraft and parts
\item \verb{[360]} ship and boat building and repairing
\item \verb{[361]} railroad locomotives and equipment
\item \verb{[362]} guided missiles, space vehicles, and parts
\item \verb{[370]} cycles and miscellaneous transportation equipment
\item \verb{[371]} scientific and controlling instruments
\item \verb{[372]} optical and health services supplies
\item \verb{[380]} photographic equipment and supplies
\item \verb{[381]} watches, clocks, and clockwork operated devices
\item \verb{[382]} not specified professional equipment
\item \verb{[390]} toys, amusement, and sporting goods
\item \verb{[391]} miscellaneous manufacturing industries
\item \verb{[392]} not specified manufacturing industries
\item \verb{[400]} railroads
\item \verb{[401]} bus service and urban transit
\item \verb{[402]} taxicab service
\item \verb{[410]} trucking service
\item \verb{[411]} warehousing and storage
\item \verb{[412]} u.s. postal service
\item \verb{[420]} water transportation
\item \verb{[421]} air transportation
\item \verb{[422]} pipe lines, except natural gas
\item \verb{[432]} services incidental to transportation
\item \verb{[440]} radio and television broadcasting
\item \verb{[441]} telephone (wire and radio)
\item \verb{[442]} telegraph and miscellaneous communication service
\item \verb{[460]} electric light and power
\item \verb{[461]} gas and steam supply systems
\item \verb{[462]} electric and gas, and other combinations
\item \verb{[470]} water supply and irrigation
\item \verb{[471]} sanitary services
\item \verb{[472]} not specified utilities
\item \verb{[500]} motor vehicles and equipment
\item \verb{[501]} furniture and home furnishings
\item \verb{[502]} lumber and construction materials
\item \verb{[510]} sporting goods, toys, and hobby goods
\item \verb{[511]} metals and minerals, except petroleum
\item \verb{[512]} electrical goods
\item \verb{[521]} hardware, plumbing and heating supplies
\item \verb{[522]} not specified electrical and hardware products
\item \verb{[530]} machinery, equipment, and supplies
\item \verb{[531]} scrap and waste materials
\item \verb{[532]} miscellaneous wholesale, durable goods
\item \verb{[540]} paper and paper products
\item \verb{[541]} drugs, chemicals, and allied products
\item \verb{[542]} apparel, fabrics, and notions
\item \verb{[550]} groceries and related products
\item \verb{[551]} farm products-raw materials
\item \verb{[552]} petroleum products
\item \verb{[560]} alcoholic beverages
\item \verb{[561]} farm supplies
\item \verb{[562]} miscellaneous wholesale, nondurable goods
\item \verb{[571]} not specified wholesale trade
\item \verb{[580]} lumber and building material retailing
\item \verb{[581]} hardware stores
\item \verb{[582]} retail nurseries and garden stores
\item \verb{[590]} mobile home dealers
\item \verb{[591]} department stores
\item \verb{[592]} variety stores
\item \verb{[600]} miscellaneous general merchandise stores
\item \verb{[601]} grocery stores
\item \verb{[602]} dairy products stores
\item \verb{[610]} retail bakeries
\item \verb{[611]} food stores, n.e.c.
\item \verb{[612]} motor vehicle dealers
\item \verb{[620]} auto and home supply stores
\item \verb{[621]} gasoline service stations
\item \verb{[622]} miscellaneous vehicle dealers
\item \verb{[630]} apparel and accessory stores, except shoe
\item \verb{[631]} shoe stores
\item \verb{[632]} furniture and home furnishings stores
\item \verb{[640]} household appliances, tv, and radio stores
\item \verb{[641]} eating and drinking places
\item \verb{[642]} drug stores
\item \verb{[650]} liquor stores
\item \verb{[651]} sporting goods, bicycles, and hobby stores
\item \verb{[652]} book and stationery stores
\item \verb{[660]} jewelry stores
\item \verb{[661]} sewing, needlework, and piece goods stores
\item \verb{[662]} mail order houses
\item \verb{[670]} vending machine operators
\item \verb{[671]} direct selling establishments
\item \verb{[672]} fuel and ice dealers
\item \verb{[681]} retail florists
\item \verb{[682]} miscellaneous retail stores
\item \verb{[691]} not specified retail trade
\item \verb{[700]} banking
\item \verb{[701]} savings and loan associations
\item \verb{[702]} credit agencies, n.e.c.
\item \verb{[710]} security, commodity brokerage, and investment companies
\item \verb{[711]} insurance
\item \verb{[712]} real estate, including real estate-insurance-law offices
\item \verb{[721]} advertising
\item \verb{[722]} services to dwellings and other buildings
\item \verb{[730]} commercial research, development, and testing labs
\item \verb{[731]} personnel supply services
\item \verb{[732]} business management and consulting services
\item \verb{[740]} computer and data processing services
\item \verb{[741]} detective and protective services
\item \verb{[742]} business services, n.e.c.
\item \verb{[750]} automotive services, except repair
\item \verb{[751]} automotive repair shops
\item \verb{[752]} electrical repair shops
\item \verb{[760]} miscellaneous repair services
\item \verb{[761]} private households
\item \verb{[762]} hotels and motels
\item \verb{[770]} lodging places, except hotels and motels
\item \verb{[771]} laundry, cleaning, and garment services
\item \verb{[772]} beauty shops
\item \verb{[780]} barber shops
\item \verb{[781]} funeral service and crematories
\item \verb{[782]} shoe repair shops
\item \verb{[790]} dressmaking shops
\item \verb{[791]} miscellaneous personal services
\item \verb{[800]} theaters and motion pictures
\item \verb{[801]} bowling alleys, billiard and pool parlors
\item \verb{[802]} miscellaneous entertainment and recreation services
\item \verb{[812]} offices of physicians
\item \verb{[820]} offices of dentists
\item \verb{[821]} offices of chiropractors
\item \verb{[822]} offices of optometrists
\item \verb{[830]} offices of health practitioners, n.e.c.
\item \verb{[831]} hospitals
\item \verb{[832]} nursing and personal care facilities
\item \verb{[840]} health services, n.e.c.
\item \verb{[841]} legal services
\item \verb{[842]} elementary and secondary schools
\item \verb{[850]} colleges and universities
\item \verb{[851]} business, trade, and vocational schools
\item \verb{[852]} libraries
\item \verb{[860]} educational services, n.e.c.
\item \verb{[861]} job training and vocational rehabilitation services
\item \verb{[862]} child day care services
\item \verb{[870]} residential care facilities, without nursing
\item \verb{[871]} social services, n.e.c.
\item \verb{[872]} museums, art galleries and zoos
\item \verb{[880]} religious organizations
\item \verb{[881]} membership organizations
\item \verb{[882]} engineering, architectural, and surveying services
\item \verb{[890]} accounting, auditing, and bookkeeping services
\item \verb{[891]} noncommercial educational and scientific research
\item \verb{[892]} miscellaneous professional and related services
\item \verb{[900]} executive and legislative offices
\item \verb{[901]} general government, n.e.c.
\item \verb{[910]} justice, public order, and safety
\item \verb{[921]} public finance, taxation, and monetary policy
\item \verb{[922]} administration of human resources programs
\item \verb{[930]} administration of environmental quality and household programs
\item \verb{[931]} administration of economic programs
\item \verb{[932]} national security and international affairs
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
}
}

\seealso{
Other Core: 
\code{\link{commute}},
\code{\link{found}},
\code{\link{indus07}},
\code{\link{industry}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{prestg80}}
}
\concept{Core}
\keyword{variable}
