% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{indusgen}
\alias{indusgen}
\title{Indust air pollution danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
indusgen
}
\details{
Question A. In general, do you think that air pollution caused by industry is…
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{chemgen}},
\code{\link{tempgen}},
\code{\link{watergen}}

Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
