% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{indusgen1}
\alias{indusgen1}
\title{In general, do you think that air pollution caused by industry is}
\source{
General Social Survey https://gss.norc.org
}
\description{
indusgen1
}
\details{
Question In general, do you think that air pollution caused by industry is…
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\keyword{variable}
