% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{inequal6}
\alias{inequal6}
\title{Everyone benefits from business profits}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal6
}
\details{
Question Do you agree or disagree . . .
F. Allowing business to make good profits is the best way to improve everyone''s standard of living.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7551/vshow}{GSS Data Explorer page for \code{inequal6}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 156 \tab 554 \tab 345 \tab 359 \tab 63 \tab 57 \tab 30 \tab - \tab 1819 \cr
   1996 \tab 1460 \tab 263 \tab 508 \tab 165 \tab 300 \tab 147 \tab 54 \tab 7 \tab - \tab 2904 \cr
   2021 \tab 2245 \tab 184 \tab 627 \tab 555 \tab 246 \tab 78 \tab 69 \tab 3 \tab 25 \tab 4032 \cr
   Total \tab 3960 \tab 603 \tab 1689 \tab 1065 \tab 905 \tab 288 \tab 180 \tab 40 \tab 25 \tab 8755 \cr
}
}

\keyword{variable}
