% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_469.R
\name{inequal7}
\alias{inequal7}
\title{Inequality due to lack of solidarity}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal7
}
\details{
Question Do you agree or disagree . . .
G. Inequality continues to exist because ordinary people don''t join together to get rid of it.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4262/vshow}{GSS Data Explorer page for \code{inequal7}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1987 \tab 255 \tab 159 \tab 534 \tab 375 \tab 328 \tab 56 \tab 79 \tab 33 \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 125 \tab 410 \tab 319 \tab 213 \tab 73 \tab 51 \tab 81 \tab - \tab 2817 \cr
   Total \tab 1800 \tab 284 \tab 944 \tab 694 \tab 541 \tab 129 \tab 130 \tab 114 \tab 0 \tab 4636 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{books16}},
\code{\link{conage}},
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{goodlife}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{justpay}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{occmobil}},
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oppol}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}},
\code{\link{opsex}},
\code{\link{opwlth}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{pasup}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychild}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payexec}},
\code{\link{payfam1}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payhard}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payresp}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{paysup}},
\code{\link{payunskl}},
\code{\link{rank}()},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}},
\code{\link{unsklrnk}}

Other Inequality, Income: 
\code{\link{eqincome}},
\code{\link{equalize}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{helppoor}},
\code{\link{incentiv}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{opoutcme}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{richeduc}},
\code{\link{richhlth}},
\code{\link{richpoor}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}}
}
\concept{ISSP Social Inequality}
\concept{Inequality, Income}
\keyword{variable}
