% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{inffilms}
\alias{inffilms}
\title{Influence of movies}
\source{
General Social Survey https://gss.norc.org
}
\description{
inffilms
}
\details{
Question Children are exposed to many influences in their daily lives. What kind of influence does each of the following have on children: Very positive, positive, neutral, negative, very negative?:
A. Movies.
}
\section{Values}{

\itemize{
\item \verb{[1]} very positive
\item \verb{[2]} positive
\item \verb{[3]} neutral
\item \verb{[4]} negative
\item \verb{[5]} very negative
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5117/vshow}{GSS Data Explorer page for \code{inffilms}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very positive \tab positive \tab neutral \tab negative \tab very negative \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 93 \tab 317 \tab 354 \tab 416 \tab 142 \tab 42 \tab 8 \tab 1372 \cr
   Total \tab 0 \tab 93 \tab 317 \tab 354 \tab 416 \tab 142 \tab 42 \tab 8 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{sppregnt}}
}
\concept{Children}
\keyword{variable}
