% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_442.R
\name{infljobs}
\alias{infljobs}
\title{Top priority of govt: inflation or jobs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
infljobs
}
\details{
Question If the government had to choose between keeping down inflation or keeping down unemployment to which do you think it should give highest priority?
}
\section{Values}{

\itemize{
\item \verb{[1]} keep down inflation
\item \verb{[2]} keep down unemplymnt
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1996 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4067/vshow}{GSS Data Explorer page for \code{infljobs}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab keep down inflation \tab keep down unemplymnt \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 275 \tab 290 \tab 103 \tab 9 \tab 1534 \cr
   1990 \tab 155 \tab 512 \tab 440 \tab 221 \tab 44 \tab 1372 \cr
   1996 \tab 1572 \tab 528 \tab 596 \tab 177 \tab 31 \tab 2904 \cr
   Total \tab 2584 \tab 1315 \tab 1326 \tab 501 \tab 84 \tab 5810 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bustax}},
\code{\link{inctax}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
