% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_133.R
\name{influasn}
\alias{influasn}
\title{Influence of asians}
\source{
General Social Survey https://gss.norc.org
}
\description{
influasn
}
\details{
Question Some people think that certain groups have too much influence in American life and politics, while other people feel that certain groups don't have as much influence as they deserve. On this card are three statements about how much influence a group might have. (1. Too much influence, 2. Just about the right amount of influence, 3. Too little influence) For each group I read to you, just tell  me the number of the statement that best says how you feel.
D. Asian Americans
}
\section{Values}{

\itemize{
\item \verb{[1]} too much influence
\item \verb{[2]} right amount
\item \verb{[3]} too little influence
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1240/vshow}{GSS Data Explorer page for \code{influasn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too much influence \tab right amount \tab too little influence \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab - \tab 85 \tab 553 \tab 503 \tab 207 \tab 24 \tab - \tab 1372 \cr
   Total \tab 0 \tab 85 \tab 553 \tab 503 \tab 207 \tab 24 \tab 0 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Asian Americans: 
\code{\link{acqasian}},
\code{\link{acqfmasn}},
\code{\link{acqnhasn}},
\code{\link{acqvaasn}},
\code{\link{acqwkasn}},
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asngovt}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{comasn}},
\code{\link{fairasns}},
\code{\link{famasns}},
\code{\link{fareasns}},
\code{\link{feelasns}},
\code{\link{hspasn10}},
\code{\link{intlasns}},
\code{\link{knwasn}},
\code{\link{liveasns}},
\code{\link{marasian}},
\code{\link{patrasns}},
\code{\link{trtasian}},
\code{\link{usasn}},
\code{\link{violasns}},
\code{\link{wlthasns}},
\code{\link{workasns}}
}
\concept{Asian Americans}
\concept{Intergroup Relations}
\keyword{variable}
