% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_132.R
\name{influblk}
\alias{influblk}
\title{Influence of blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
influblk
}
\details{
Question Some people think that certain groups have too much influence in American life and politics, while other people feel that certain groups don't have as much influence as they deserve. On this card are three statements about how much influence a group might have. (1. Too much influence, 2. Just about the right amount of influence, 3. Too little influence) For each group I read to you, just tell  me the number of the statement that best says how you feel.
C. Blacks
}
\section{Values}{

\itemize{
\item \code{1} too much influence
\item \code{2} right amount
\item \code{3} too little influence
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1239/vshow}{GSS Data Explorer page for \code{influblk}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{blkcol}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{influjew}},
\code{\link{influwht}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}}
}
\concept{Blacks}
\concept{Intergroup Relations}
\keyword{variable}
