% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_132.R
\name{influjew}
\alias{influjew}
\title{Influence of jews}
\source{
General Social Survey https://gss.norc.org
}
\description{
influjew
}
\details{
Question Some people think that certain groups have too much influence in American life and politics, while other people feel that certain groups don't have as much influence as they deserve. On this card are three statements about how much influence a group might have. (1. Too much influence, 2. Just about the right amount of influence, 3. Too little influence) For each group I read to you, just tell  me the number of the statement that best says how you feel.
B. Jews
}
\section{Values}{

\itemize{
\item \verb{[1]} too much influence
\item \verb{[2]} right amount
\item \verb{[3]} too little influence
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1238/vshow}{GSS Data Explorer page for \code{influjew}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too much influence \tab right amount \tab too little influence \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab - \tab 286 \tab 735 \tab 170 \tab 158 \tab 23 \tab - \tab 1372 \cr
   Total \tab 0 \tab 286 \tab 735 \tab 170 \tab 158 \tab 23 \tab 0 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Jews: 
\code{\link{bmitzvah}},
\code{\link{comjews}},
\code{\link{fairjews}},
\code{\link{famjews}},
\code{\link{farejews}},
\code{\link{intljews}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{jewtemp}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{knwjew}},
\code{\link{livejews}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{marjew}},
\code{\link{pajew}},
\code{\link{pajwoth}},
\code{\link{patrjews}},
\code{\link{spjewaj}},
\code{\link{spjoth16}},
\code{\link{usjews}},
\code{\link{violjews}},
\code{\link{wlthjews}},
\code{\link{workjews}}
}
\concept{Intergroup Relations}
\concept{Jews}
\keyword{variable}
