% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_132.R
\name{influwht}
\alias{influwht}
\title{Influence of whites}
\source{
General Social Survey https://gss.norc.org
}
\description{
influwht
}
\details{
Question Some people think that certain groups have too much influence in American life and politics, while other people feel that certain groups don't have as much influence as they deserve. On this card are three statements about how much influence a group might have. (1. Too much influence, 2. Just about the right amount of influence, 3. Too little influence) For each group I read to you, just tell  me the number of the statement that best says how you feel.
A. Whites
}
\section{Values}{

\itemize{
\item \verb{[1]} too much influence
\item \verb{[2]} right amount
\item \verb{[3]} too little influence
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1237/vshow}{GSS Data Explorer page for \code{influwht}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too much influence \tab right amount \tab too little influence \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 340 \tab 867 \tab 78 \tab 65 \tab 22 \tab 1372 \cr
   Total \tab 0 \tab 340 \tab 867 \tab 78 \tab 65 \tab 22 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{blkcol}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{influblk}},
\code{\link{influjew}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}}
}
\concept{Intergroup Relations}
\keyword{variable}
