% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{infnettv}
\alias{infnettv}
\title{Influence of programs on network tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
infnettv
}
\details{
Question Children are exposed to many influences in their daily lives. What kind of influence does each of the following have on children: Very positive, positive, neutral, negative, very negative?:
C. Programs on network television.
}
\section{Values}{

\itemize{
\item \code{1} very positive
\item \code{2} positive
\item \code{3} neutral
\item \code{4} negative
\item \code{5} very negative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5119/vshow}{GSS Data Explorer page for \code{infnettv}}.
}

\seealso{
Other Children: 
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infpubtv}},
\code{\link{sppregnt}}
}
\concept{Children}
\keyword{variable}
