% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{infobiz}
\alias{infobiz}
\title{R trust business for pollution information}
\source{
General Social Survey https://gss.norc.org
}
\description{
infobiz
}
\details{
Question How much trust do you have in each of the following groups to give you correct information about causes of pollution?
A. Business and industry
}
\section{Values}{

\itemize{
\item \code{1} a great deal of trust
\item \code{2} quite a lot of trust
\item \code{3} some trust
\item \code{4} not much trust
\item \code{5} hardly any trust
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4791/vshow}{GSS Data Explorer page for \code{infobiz}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a great deal of trust \tab quite a lot of trust \tab some trust \tab not much trust \tab hardly any trust \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 34 \tab 82 \tab 417 \tab 415 \tab 184 \tab 39 \tab 105 \tab 2817 \cr
   Total \tab 1541 \tab 34 \tab 82 \tab 417 \tab 415 \tab 184 \tab 39 \tab 105 \tab 2817 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{econgrn}},
\code{\link{excldimm}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{ldcgrn}},
\code{\link{nukeacc}}
}
\concept{ISSP Environment}
\keyword{variable}
