% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{infoprofit}
\alias{infoprofit}
\title{Would r provide info companies could profit from}
\source{
General Social Survey https://gss.norc.org
}
\description{
infoprofit
}
\details{
Question Private companies offer services in return for personal information such as your name, phone number, email address, online-search history, or purchases. Regardless of whether you use the Internet yourself, please indicate the extent to which you agree or disagree with the following statements.… I would provide my personal information online even though private companies may use it to make a profit.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8053/vshow}{GSS Data Explorer page for \code{infoprofit}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 19 \tab 123 \tab 265 \tab 515 \tab 619 \tab 76 \tab 1673 \tab 7 \tab 12 \tab 3309 \cr
   Total \tab 0 \tab 19 \tab 123 \tab 265 \tab 515 \tab 619 \tab 76 \tab 1673 \tab 7 \tab 12 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{dataprot}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{modeorg}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
