% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{infotv}
\alias{infotv}
\title{R trust radio or tv for pollution information}
\source{
General Social Survey https://gss.norc.org
}
\description{
infotv
}
\details{
Question How much trust do you have in each of the following groups to give you correct information about causes of pollution?
E. Radio or TV programs
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal of trust
\item \verb{[2]} quite a lot of trust
\item \verb{[3]} some trust
\item \verb{[4]} not much trust
\item \verb{[5]} hardly any trust
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4795/vshow}{GSS Data Explorer page for \code{infotv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a great deal of trust \tab quite a lot of trust \tab some trust \tab not much trust \tab hardly any trust \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 51 \tab 229 \tab 559 \tab 215 \tab 72 \tab 33 \tab 117 \tab 2817 \cr
   Total \tab 1541 \tab 51 \tab 229 \tab 559 \tab 215 \tab 72 \tab 33 \tab 117 \tab 2817 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{econgrn}},
\code{\link{excldimm}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{ldcgrn}},
\code{\link{nukeacc}}
}
\concept{ISSP Environment}
\keyword{variable}
