% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{infpubtv}
\alias{infpubtv}
\title{Influence of children's programs on public tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
infpubtv
}
\details{
Question Children are exposed to many influences in their daily lives. What kind of influence does each of the following have on children: Very positive, positive, neutral, negative, very negative?:
B. Children''s programs on television.
}
\section{Values}{

\itemize{
\item \code{1} very positive
\item \code{2} positive
\item \code{3} neutral
\item \code{4} negative
\item \code{5} very negative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5118/vshow}{GSS Data Explorer page for \code{infpubtv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very positive \tab positive \tab neutral \tab negative \tab very negative \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 299 \tab 811 \tab 146 \tab 54 \tab 17 \tab 38 \tab 7 \tab 1372 \cr
   Total \tab 0 \tab 299 \tab 811 \tab 146 \tab 54 \tab 17 \tab 38 \tab 7 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{sppregnt}}
}
\concept{Children}
\keyword{variable}
