% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{infrmgrp}
\alias{infrmgrp}
\title{Any activities with the same group of ppl regularly}
\source{
General Social Survey https://gss.norc.org
}
\description{
infrmgrp
}
\details{
Question Are there any activities that you do with the same group of people on a regular basis even if the group doesn't have a name such as a bridge group, exercise group, or a group that meets to discuss individual or community problems?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3109/vshow}{GSS Data Explorer page for \code{infrmgrp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 394 \tab 1071 \tab 1 \tab 6 \tab 2812 \cr
   Total \tab 1340 \tab 394 \tab 1071 \tab 1 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{churchtx}},
\code{\link{typfrat}},
\code{\link{typpolit}},
\code{\link{typserv}},
\code{\link{typvet}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
