% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{inlawnum}
\alias{inlawnum}
\title{Number of in-laws}
\source{
General Social Survey https://gss.norc.org
}
\description{
inlawnum
}
\details{
Question Now thinking of all your adult relatives -- those still living and age 18 or older.
How many of each do you have?  READ EACH ITEM. CODE A NUMBER OR ZERO FOR EACH.
D. Parents-in-Law and Adult Brothers- and Sisters-in-Law
}
\section{Values}{

\itemize{
\item \code{96} 96 or more
\item \code{97} # dk but ) 0
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4156/vshow}{GSS Data Explorer page for \code{inlawnum}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 22 \tab 23 \tab 24 \tab 29 \tab 3 \tab 30 \tab 35 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab # dk but ) 0 \tab don't know \tab no answer \tab Total \cr
   1986 \tab - \tab 305 \tab 161 \tab 52 \tab 21 \tab 18 \tab 11 \tab 5 \tab 7 \tab 11 \tab 3 \tab 1 \tab 3 \tab 169 \tab 10 \tab 1 \tab 1 \tab 4 \tab 1 \tab 133 \tab 1 \tab 1 \tab 116 \tab 127 \tab 130 \tab 64 \tab 65 \tab 36 \tab 5 \tab 4 \tab 4 \tab 1470 \cr
   Total \tab 0 \tab 305 \tab 161 \tab 52 \tab 21 \tab 18 \tab 11 \tab 5 \tab 7 \tab 11 \tab 3 \tab 1 \tab 3 \tab 169 \tab 10 \tab 1 \tab 1 \tab 4 \tab 1 \tab 133 \tab 1 \tab 1 \tab 116 \tab 127 \tab 130 \tab 64 \tab 65 \tab 36 \tab 5 \tab 4 \tab 4 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
