% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{instype01}
\alias{instype01}
\title{First type of insurance mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
instype01
}
\details{
Question What kind of health insurance do you have?  Please check all that apply.  First type of insurance mentioned.
}
\section{Values}{

\itemize{
\item \code{1} public health insurance
\item \code{2} private insurance
\item \code{3} employer/union based insurance
\item \code{4} other
\item \code{5} have no health insurance
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7972/vshow}{GSS Data Explorer page for \code{instype01}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab public health insurance \tab private insurance \tab employer/union based insurance \tab other \tab have no health insurance \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2545 \tab 363 \tab 153 \tab 358 \tab 36 \tab 71 \tab 1 \tab 8 \tab 9 \tab 3544 \cr
   Total \tab 2545 \tab 363 \tab 153 \tab 358 \tab 36 \tab 71 \tab 1 \tab 8 \tab 9 \tab 3544 \cr
}
}

\keyword{variable}
