% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{instype04}
\alias{instype04}
\title{Fourth type of insurance mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
instype04
}
\details{
Question What kind of health insurance do you have?  Please check all that apply.  Fourth type of insurance mentioned.
}
\section{Values}{

\itemize{
\item \verb{[1]} public health insurance
\item \verb{[2]} private insurance
\item \verb{[3]} employer/union based insurance
\item \verb{[4]} other
\item \verb{[5]} have no health insurance
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7975/vshow}{GSS Data Explorer page for \code{instype04}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab other \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3525 \tab 1 \tab 1 \tab 8 \tab 9 \tab 3544 \cr
   Total \tab 3525 \tab 1 \tab 1 \tab 8 \tab 9 \tab 3544 \cr
}
}

\keyword{variable}
