% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_673.R
\name{intaccprxy}
\alias{intaccprxy}
\title{How often r asked someone to use the internet on their behalf}
\source{
General Social Survey https://gss.norc.org
}
\description{
intaccprxy
}
\details{
Question In the last 12 months how often, if at all, did you ask someone to do something for you on the Internet on your behalf, such as sending an email, getting information or shopping online?
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[2]} once or twice
\item \verb{[3]} more often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8114/vshow}{GSS Data Explorer page for \code{intaccprxy}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab not at all \tab once or twice \tab more often \tab can't choose \tab iap \tab Total \cr
   2024 \tab - \tab 29 \tab 18 \tab 9 \tab 8 \tab 3245 \tab 3309 \cr
   Total \tab 0 \tab 29 \tab 18 \tab 9 \tab 8 \tab 3245 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{ftruse}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint6}},
\code{\link{nouseint7}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{prxyuse}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
