% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{intage}
\alias{intage}
\title{Age of interviewer}
\source{
General Social Survey https://gss.norc.org
}
\description{
intage
}
\details{
Question A. Age of interviewer
}
\section{Values}{

\itemize{
\item \verb{[1]} 18-34
\item \verb{[2]} 35-44
\item \verb{[3]} 45-54
\item \verb{[4]} 55-64
\item \verb{[5]} 65 or older
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5563/vshow}{GSS Data Explorer page for \code{intage}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab 18-34 \tab 35-44 \tab 45-54 \tab 55-64 \tab 65 or older \tab no answer \tab iap \tab Total \cr
   2004 \tab - \tab 157 \tab 586 \tab 772 \tab 873 \tab 354 \tab 70 \tab - \tab 2812 \cr
   2006 \tab - \tab 142 \tab 634 \tab 1516 \tab 1652 \tab 566 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 151 \tab 204 \tab 732 \tab 720 \tab 212 \tab 4 \tab - \tab 2023 \cr
   2010 \tab - \tab 171 \tab 285 \tab 538 \tab 710 \tab 313 \tab 27 \tab - \tab 2044 \cr
   2012 \tab - \tab 79 \tab 177 \tab 551 \tab 721 \tab 446 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 55 \tab 267 \tab 736 \tab 1004 \tab 476 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 159 \tab 362 \tab 473 \tab 1194 \tab 679 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 138 \tab 295 \tab 539 \tab 750 \tab 626 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   Total \tab 4032 \tab 1052 \tab 2810 \tab 5857 \tab 7624 \tab 3672 \tab 101 \tab 0 \tab 25148 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}},
\code{\link{mode}()}

Other Interview: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}},
\code{\link{mode}()}
}
\concept{Interview}
\concept{Paradata}
\keyword{variable}
