% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intapps}
\alias{intapps}
\title{R's skill at learning new software}
\source{
General Social Survey https://gss.norc.org
}
\description{
intapps
}
\details{
Question How would you rate your ability to do the following?… Learning how to use new software or apps?
}
\section{Values}{

\itemize{
\item \verb{[1]} very good
\item \verb{[2]} good
\item \verb{[3]} neither good nor bad
\item \verb{[4]} bad
\item \verb{[5]} very bad
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8147/vshow}{GSS Data Explorer page for \code{intapps}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab very good \tab good \tab neither good nor bad \tab bad \tab very bad \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 450 \tab 530 \tab 337 \tab 145 \tab 74 \tab 27 \tab 1732 \tab 7 \tab 7 \tab 3309 \cr
   Total \tab 0 \tab 450 \tab 530 \tab 337 \tab 145 \tab 74 \tab 27 \tab 1732 \tab 7 \tab 7 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intbhlf}},
\code{\link{inthome}},
\code{\link{intliteracy}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplgrsn}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
