% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intbhlf}
\alias{intbhlf}
\title{Frequency r was asked to use internet for someone else}
\source{
General Social Survey https://gss.norc.org
}
\description{
intbhlf
}
\details{
Question In the last 12 months, how often, if at all, were you asked by someone to do something on their behalf on the Internet, such as sending an email, getting information or shopping online? Exclude work related tasks.
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[2]} once or twice
\item \verb{[3]} more often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8115/vshow}{GSS Data Explorer page for \code{intbhlf}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab not at all \tab once or twice \tab more often \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 557 \tab 502 \tab 433 \tab 74 \tab 1731 \tab 4 \tab 8 \tab 3309 \cr
   Total \tab 0 \tab 557 \tab 502 \tab 433 \tab 74 \tab 1731 \tab 4 \tab 8 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{inthome}},
\code{\link{intliteracy}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplgrsn}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
