% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{intcnfrm}
\alias{intcnfrm}
\title{How often r tries to confirm information with other sources}
\source{
General Social Survey https://gss.norc.org
}
\description{
intcnfrm
}
\details{
Question When looking for news or information about political matters online, how often, if at all, do you… Try to confirm information you find by searching online for another source?
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} rarely
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8182/vshow}{GSS Data Explorer page for \code{intcnfrm}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab very often \tab often \tab sometimes \tab rarely \tab never \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 283 \tab 477 \tab 482 \tab 162 \tab 111 \tab 44 \tab 1732 \tab 7 \tab 11 \tab 3309 \cr
   Total \tab 0 \tab 283 \tab 477 \tab 482 \tab 162 \tab 111 \tab 44 \tab 1732 \tab 7 \tab 11 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{inthrss}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intrust}},
\code{\link{intscam}},
\code{\link{intviews}},
\code{\link{nextgen1}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
