% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{intcntct}
\alias{intcntct}
\title{How much of r's communication is via text, mobile phone, or internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
intcntct
}
\details{
Question Think now of your contact with all of your family members and close friends. How much of it is through text messages, mobile phones,
or other communication devices that use the internet?
}
\section{Values}{

\itemize{
\item \code{1} all or almost all of it
\item \code{2} most of it
\item \code{3} about half of it
\item \code{4} some of it
\item \code{5} none or almost none of it
\item \code{6} i do not use any of these devices
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7270/vshow}{GSS Data Explorer page for \code{intcntct}}.
}

\keyword{variable}
