% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_675.R
\name{intcomm}
\alias{intcomm}
\title{How often r uses internet for communication}
\source{
General Social Survey https://gss.norc.org
}
\description{
intcomm
}
\details{
Question How would you rate your ability to do the following?… Communicating (for example writing, talking, video calling)?
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} hardly ever
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8116/vshow}{GSS Data Explorer page for \code{intcomm}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab very often \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 697 \tab 432 \tab 272 \tab 88 \tab 52 \tab 20 \tab 1732 \tab 5 \tab 11 \tab 3309 \cr
   Total \tab 0 \tab 697 \tab 432 \tab 272 \tab 88 \tab 52 \tab 20 \tab 1732 \tab 5 \tab 11 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{intlnly}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intsearch}},
\code{\link{intshare}},
\code{\link{intstream}},
\code{\link{inttrig}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
