% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{intecon}
\alias{intecon}
\title{Interested in economic issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
intecon
}
\details{
Question Economic issues and business conditions. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} moderately interested
\item \code{3} not at all interested
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3456/vshow}{GSS Data Explorer page for \code{intecon}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 518 \tab 748 \tab 594 \tab 157 \tab 5 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1334 \tab 319 \tab 301 \tab 83 \tab 1 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 508 \tab 376 \tab 103 \tab 11 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 530 \tab 524 \tab 180 \tab 4 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 546 \tab 638 \tab 201 \tab 2 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 476 \tab 524 \tab 174 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 2145 \tab 812 \tab 891 \tab 164 \tab - \tab - \tab 20 \tab 4032 \cr
   Total \tab 8918 \tab 3939 \tab 3848 \tab 1062 \tab 24 \tab 15 \tab 20 \tab 17826 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{scimode}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
