% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{inteduc}
\alias{inteduc}
\title{Interested in local school issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
inteduc
}
\details{
Question Local school issues. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} moderately interested
\item \verb{[3]} not at all interested
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3454/vshow}{GSS Data Explorer page for \code{inteduc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 518 \tab 709 \tab 592 \tab 198 \tab 5 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1334 \tab 390 \tab 249 \tab 64 \tab 1 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 541 \tab 348 \tab 103 \tab 5 \tab 5 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 593 \tab 479 \tab 166 \tab - \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 605 \tab 592 \tab 189 \tab - \tab 4 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 570 \tab 449 \tab 156 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 2143 \tab 541 \tab 930 \tab 407 \tab 2 \tab - \tab 9 \tab 4032 \cr
   Total \tab 8916 \tab 3949 \tab 3639 \tab 1283 \tab 13 \tab 17 \tab 9 \tab 17826 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{scimode}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
