% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_382.R
\name{intenvir}
\alias{intenvir}
\title{Interested in environmental issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
intenvir
}
\details{
Question Issues about environmental pollution. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} moderately interested
\item \verb{[3]} not at all interested
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3460/vshow}{GSS Data Explorer page for \code{intenvir}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 751 \tab 600 \tab 142 \tab 11 \tab 1 \tab 2023 \cr
   2010 \tab 1334 \tab 315 \tab 330 \tab 57 \tab 2 \tab 6 \tab 2044 \cr
   2012 \tab 972 \tab 467 \tab 430 \tab 94 \tab 7 \tab 4 \tab 1974 \cr
   2014 \tab 1299 \tab 510 \tab 598 \tab 130 \tab - \tab 1 \tab 2538 \cr
   2016 \tab 1477 \tab 586 \tab 651 \tab 149 \tab 1 \tab 3 \tab 2867 \cr
   2018 \tab 1173 \tab 531 \tab 516 \tab 126 \tab 2 \tab - \tab 2348 \cr
   Total \tab 6773 \tab 3160 \tab 3125 \tab 698 \tab 23 \tab 15 \tab 13794 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intmil}},
\code{\link{intspace}},
\code{\link{sciintro}},
\code{\link{scimath}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
