% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_55.R
\name{interest}
\alias{interest}
\title{Interested how and why things happen}
\source{
General Social Survey https://gss.norc.org
}
\description{
interest
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
12. That he is interested in how and why things happen
}
\section{Values}{

\itemize{
\item \code{1} 1 most desirable
\item \code{2} 3 most desirable
\item \code{3} not mentioned
\item \code{4} 3 least desirable
\item \code{5} 1 least desirable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983 \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/485/vshow}{GSS Data Explorer page for \code{interest}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 62 \tab 230 \tab 681 \tab 323 \tab 204 \tab 4 \tab - \tab 1504 \cr
   1975 \tab - \tab 42 \tab 165 \tab 659 \tab 396 \tab 218 \tab 10 \tab - \tab 1490 \cr
   1976 \tab - \tab 47 \tab 234 \tab 692 \tab 341 \tab 175 \tab 10 \tab - \tab 1499 \cr
   1978 \tab - \tab 41 \tab 176 \tab 660 \tab 426 \tab 217 \tab 10 \tab 2 \tab 1532 \cr
   1980 \tab 975 \tab 20 \tab 63 \tab 219 \tab 125 \tab 62 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 33 \tab 176 \tab 648 \tab 465 \tab 257 \tab 20 \tab - \tab 1599 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   Total \tab 3019 \tab 245 \tab 1044 \tab 3559 \tab 2076 \tab 1133 \tab 57 \tab 3 \tab 11136 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicably}},
\code{\link{cleany}},
\code{\link{controly}},
\code{\link{honesty}},
\code{\link{judgmeny}},
\code{\link{mannersy}},
\code{\link{roley}},
\code{\link{studious}},
\code{\link{successy}}
}
\concept{Core}
\keyword{variable}
