% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_57.R
\name{interesz}
\alias{interesz}
\title{Interested how things happen-version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
interesz
}
\details{
Question Please look at the qualities listed on this card. (SHOW CARD:
LET R. INSPECT) All of the qualities may be desirable for a child to have, but could you tell me whether the quality is extremely important, very important, fairly important, not too important, or not at all important. READ AND RECORD RESPONSES FOR EACH.
L. That he is interested in how and why things happen
}
\section{Values}{

\itemize{
\item \code{1} extremely impt
\item \code{2} very impt
\item \code{3} fairly impt
\item \code{4} not too impt
\item \code{5} not at all impt
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1980 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/511/vshow}{GSS Data Explorer page for \code{interesz}}.
}

\seealso{
Other Core: 
\code{\link{amicablz}},
\code{\link{considez}},
\code{\link{controlz}},
\code{\link{obey}},
\code{\link{obeysz}},
\code{\link{popular}},
\code{\link{responsz}},
\code{\link{rolez}},
\code{\link{studiouz}}

Other Children: 
\code{\link{amicablz}},
\code{\link{considez}},
\code{\link{controlz}},
\code{\link{obey}},
\code{\link{obeysz}},
\code{\link{popular}},
\code{\link{responsz}},
\code{\link{rolez}},
\code{\link{studiouz}}

Other Split Ballots: 
\code{\link{amicablz}},
\code{\link{considez}},
\code{\link{controlz}},
\code{\link{obeysz}},
\code{\link{responsz}},
\code{\link{rolez}},
\code{\link{studiouz}}
}
\concept{Children}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
