% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_539.R
\name{interpol}
\alias{interpol}
\title{Joined an internet political forum}
\source{
General Social Survey https://gss.norc.org
}
\description{
interpol
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the past year, whether you have done it in the more distant past, whether you have not done it but might do it, or have not done it and would never, under any circumstances, do it.
H. Joined an Internet political forum or discussion group
}
\section{Values}{

\itemize{
\item \verb{[1]} have done it in the past yr
\item \verb{[2]} have done it in the more distant past
\item \verb{[3]} have not done it but might do it
\item \verb{[4]} have not done it and would never do it
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4927/vshow}{GSS Data Explorer page for \code{interpol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab have done it in the past yr \tab have done it in the more distant past \tab have not done it but might do it \tab have not done it and would never do it \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 81 \tab 28 \tab 455 \tab 903 \tab 2 \tab 3 \tab 2812 \cr
   Total \tab 1340 \tab 81 \tab 28 \tab 455 \tab 903 \tab 2 \tab 3 \tab 2812 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{cntctgov}},
\code{\link{joindem}},
\code{\link{polfunds}},
\code{\link{racmeet}},
\code{\link{revmeet}},
\code{\link{signdpet}},
\code{\link{usemedia}}
}
\concept{ISSP Citizenship}
\keyword{variable}
