% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{intethn}
\alias{intethn}
\title{Race of interviewer}
\source{
General Social Survey https://gss.norc.org
}
\description{
intethn
}
\details{
Question B. Race/ethnicity of interviewer
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black
\item \code{3} hispanic
\item \code{4} asian
\item \code{5} two or more races
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5565/vshow}{GSS Data Explorer page for \code{intethn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab white \tab black \tab hispanic \tab asian \tab no answer \tab two or more races \tab not available in this release \tab Total \cr
   2004 \tab - \tab 2255 \tab 387 \tab 74 \tab 26 \tab 70 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 3233 \tab 528 \tab 334 \tab 128 \tab 287 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1522 \tab 249 \tab 193 \tab 7 \tab 4 \tab 48 \tab - \tab 2023 \cr
   2010 \tab - \tab 1380 \tab 297 \tab 147 \tab 23 \tab 83 \tab 114 \tab - \tab 2044 \cr
   2012 \tab - \tab 1458 \tab 245 \tab 90 \tab - \tab 2 \tab 179 \tab - \tab 1974 \cr
   2014 \tab - \tab 1986 \tab 248 \tab 234 \tab 35 \tab - \tab 35 \tab - \tab 2538 \cr
   2016 \tab - \tab 1792 \tab 451 \tab 368 \tab 11 \tab - \tab 245 \tab - \tab 2867 \cr
   2018 \tab - \tab 1770 \tab 293 \tab 183 \tab 13 \tab - \tab 89 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 0 \tab 15396 \tab 2698 \tab 1623 \tab 243 \tab 446 \tab 710 \tab 4032 \tab 25148 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}},
\code{\link{mode}()}
}
\concept{Paradata}
\keyword{variable}
