% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{intfarm}
\alias{intfarm}
\title{Interested in farm issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
intfarm
}
\details{
Question Agricultural and farm issues. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} moderately interested
\item \code{3} not at all interested
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3453/vshow}{GSS Data Explorer page for \code{intfarm}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 369 \tab 770 \tab 359 \tab 6 \tab 1 \tab 2023 \cr
   2010 \tab 1334 \tab 182 \tab 389 \tab 131 \tab 2 \tab 6 \tab 2044 \cr
   2012 \tab 972 \tab 236 \tab 493 \tab 262 \tab 7 \tab 4 \tab 1974 \cr
   2014 \tab 1299 \tab 279 \tab 608 \tab 350 \tab 1 \tab 1 \tab 2538 \cr
   2016 \tab 1477 \tab 302 \tab 657 \tab 424 \tab 3 \tab 4 \tab 2867 \cr
   2018 \tab 1173 \tab 237 \tab 573 \tab 365 \tab - \tab - \tab 2348 \cr
   Total \tab 6773 \tab 1605 \tab 3490 \tab 1891 \tab 19 \tab 16 \tab 13794 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{scimode}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
