% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{inthrss}
\alias{inthrss}
\title{Has r or someone they know been harassed online}
\source{
General Social Survey https://gss.norc.org
}
\description{
inthrss
}
\details{
Question Some people report experiences of harassment on the Internet. Have you or someone you know personally ever experienced online harassment or online hate speech?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, me myself
\item \verb{[2]} yes, someone i know personally
\item \verb{[3]} yes, me and someone else i know personally
\item \verb{[4]} no
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8125/vshow}{GSS Data Explorer page for \code{inthrss}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab yes, me myself \tab yes, someone i know personally \tab yes, me and someone else i know personally \tab no \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 147 \tab 250 \tab 103 \tab 1047 \tab 77 \tab 1670 \tab 4 \tab 11 \tab 3309 \cr
   Total \tab 0 \tab 147 \tab 250 \tab 103 \tab 1047 \tab 77 \tab 1670 \tab 4 \tab 11 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{intcnfrm}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intrust}},
\code{\link{intscam}},
\code{\link{intviews}},
\code{\link{nextgen1}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
