% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{intjob}
\alias{intjob}
\title{Importance of interesting work in a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
intjob
}
\details{
Question On the following list there are various aspects of jobs. Please circle one number to show how important you personally consider it is in a job:
E. An interesting job?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} neither important nor unimportant
\item \code{4} not important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4511/vshow}{GSS Data Explorer page for \code{intjob}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab neither important nor unimportant \tab not important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 617 \tab 711 \tab 65 \tab 18 \tab 5 \tab 11 \tab 26 \tab 1537 \cr
   1998 \tab 1609 \tab 596 \tab 515 \tab 37 \tab 16 \tab 3 \tab 15 \tab 41 \tab 2832 \cr
   2006 \tab 2992 \tab 863 \tab 565 \tab 61 \tab 17 \tab 5 \tab 5 \tab 2 \tab 4510 \cr
   2016 \tab 1390 \tab 758 \tab 603 \tab 80 \tab 27 \tab 1 \tab 3 \tab 5 \tab 2867 \cr
   Total \tab 6075 \tab 2834 \tab 2394 \tab 243 \tab 78 \tab 14 \tab 34 \tab 74 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{bosswrks}},
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{hiinc}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}},
\code{\link{strngun}}
}
\concept{Work Orientation}
\keyword{variable}
