% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_128.R
\name{intlasns}
\alias{intlasns}
\title{Unintelligent - intelligent}
\source{
General Social Survey https://gss.norc.org
}
\description{
intlasns
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
D. Do people in these groups tend to be unintelligent or tend to be intelligent?
4. Asian Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} unintelligent
\item \verb{[7]} intelligent
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1199/vshow}{GSS Data Explorer page for \code{intlasns}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab unintelligent \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab intelligent \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 13 \tab 51 \tab 122 \tab 494 \tab 263 \tab 167 \tab 94 \tab 136 \tab 32 \tab 1372 \cr
   2000 \tab 1419 \tab 7 \tab 36 \tab 86 \tab 487 \tab 269 \tab 228 \tab 114 \tab 157 \tab 14 \tab 2817 \cr
   Total \tab 1419 \tab 20 \tab 87 \tab 208 \tab 981 \tab 532 \tab 395 \tab 208 \tab 293 \tab 46 \tab 4189 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{intlblks}},
\code{\link{intljews}},
\code{\link{intlwhts}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
