% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{intlcowk}
\alias{intlcowk}
\title{How often exchange info with co-workers located outside usa}
\source{
General Social Survey https://gss.norc.org
}
\description{
intlcowk
}
\details{
Question How often do you exchange information with co-workers located outside the United States? Is it…
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} once per year or less
\item \code{3} several times per year
\item \code{4} several times per month
\item \code{5} several times per week
\item \code{6} daily
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3892/vshow}{GSS Data Explorer page for \code{intlcowk}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab once per year or less \tab several times per year \tab several times per month \tab several times per week \tab daily \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 673 \tab 36 \tab 30 \tab 27 \tab 19 \tab 53 \tab 6 \tab 14 \tab 2023 \cr
   Total \tab 1165 \tab 673 \tab 36 \tab 30 \tab 27 \tab 19 \tab 53 \tab 6 \tab 14 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
