% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{intlincs}
\alias{intlincs}
\title{Large intl company damage to local business}
\source{
General Social Survey https://gss.norc.org
}
\description{
intlincs
}
\details{
Question How much do you agree or disagree with the following statements? (Please, select one response)
A. Large international companies are doing more and more damage to local businesses in America.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4883/vshow}{GSS Data Explorer page for \code{intlincs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 272 \tab 435 \tab 287 \tab 150 \tab 17 \tab 54 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 217 \tab 568 \tab 269 \tab 132 \tab 17 \tab 68 \tab 3 \tab 2538 \cr
   Total \tab 2860 \tab 489 \tab 1003 \tab 556 \tab 282 \tab 34 \tab 122 \tab 4 \tab 5350 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
