% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_675.R
\name{intlnly}
\alias{intlnly}
\title{R would feel lonely without internet?}
\source{
General Social Survey https://gss.norc.org
}
\description{
intlnly
}
\details{
Question How much do you agree or disagree with the following statements?… Without the Internet I would feel lonely.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8054/vshow}{GSS Data Explorer page for \code{intlnly}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 88 \tab 239 \tab 352 \tab 516 \tab 338 \tab 33 \tab 1732 \tab 2 \tab 9 \tab 3309 \cr
   Total \tab 0 \tab 88 \tab 239 \tab 352 \tab 516 \tab 338 \tab 33 \tab 1732 \tab 2 \tab 9 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intcomm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intsearch}},
\code{\link{intshare}},
\code{\link{intstream}},
\code{\link{inttrig}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
