% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{intmed}
\alias{intmed}
\title{Interested in medical discoveries}
\source{
General Social Survey https://gss.norc.org
}
\description{
intmed
}
\details{
Question Issues about new medical discoveries. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} moderately interested
\item \code{3} not at all interested
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3458/vshow}{GSS Data Explorer page for \code{intmed}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 907 \tab 496 \tab 93 \tab 7 \tab 2 \tab 2023 \cr
   2010 \tab 1334 \tab 420 \tab 251 \tab 32 \tab 1 \tab 6 \tab 2044 \cr
   2012 \tab 972 \tab 599 \tab 346 \tab 48 \tab 5 \tab 4 \tab 1974 \cr
   2014 \tab 1299 \tab 725 \tab 448 \tab 65 \tab - \tab 1 \tab 2538 \cr
   2016 \tab 1477 \tab 825 \tab 493 \tab 68 \tab 1 \tab 3 \tab 2867 \cr
   2018 \tab 1173 \tab 669 \tab 443 \tab 62 \tab 1 \tab - \tab 2348 \cr
   Total \tab 6773 \tab 4145 \tab 2477 \tab 368 \tab 15 \tab 16 \tab 13794 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{scimode}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
