% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{intoppo}
\alias{intoppo}
\title{Trend of online contact with politically different}
\source{
General Social Survey https://gss.norc.org
}
\description{
intoppo
}
\details{
Question In the last 12 months, if at all, has your online contact with the following people increased, decreased or remained the same?… People with different political views from your own.
}
\section{Values}{

\itemize{
\item \verb{[1]} greatly decreased
\item \verb{[2]} somewhat decreased
\item \verb{[3]} remained the same
\item \verb{[4]} somewhat increased
\item \verb{[5]} greatly increased
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8122/vshow}{GSS Data Explorer page for \code{intoppo}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab greatly decreased \tab somewhat decreased \tab remained the same \tab somewhat increased \tab greatly increased \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 70 \tab 151 \tab 948 \tab 90 \tab 15 \tab 263 \tab 1732 \tab 15 \tab 25 \tab 3309 \cr
   Total \tab 0 \tab 70 \tab 151 \tab 948 \tab 90 \tab 15 \tab 263 \tab 1732 \tab 15 \tab 25 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{intcnfrm}},
\code{\link{inthrss}},
\code{\link{intpolview}},
\code{\link{intrust}},
\code{\link{intscam}},
\code{\link{intviews}},
\code{\link{nextgen1}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
