% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{intpol}
\alias{intpol}
\title{How interested in politics is r}
\source{
General Social Survey https://gss.norc.org
}
\description{
intpol
}
\details{
Question How interested are you in politics and national affairs?  Are you very interested, somewhat interested, only slightly interested, or not at all interested?
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} somewhat int
\item \verb{[3]} slightly int
\item \verb{[4]} not interested
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1027/vshow}{GSS Data Explorer page for \code{intpol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very interested \tab somewhat int \tab slightly int \tab not interested \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 402 \tab 742 \tab 481 \tab 180 \tab 5 \tab 9 \tab 1819 \cr
   Total \tab 0 \tab 402 \tab 742 \tab 481 \tab 180 \tab 5 \tab 9 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
