% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intprfssnl}
\alias{intprfssnl}
\title{Frequency r has work/school internet access}
\source{
General Social Survey https://gss.norc.org
}
\description{
intprfssnl
}
\details{
Question How often, if at all, do you typically use the Internet on any device at the following locations for any reason?… At your work-place/at school/ at university?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} hardly ever
\item \verb{[5]} never
\item \verb{[6]} no access
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8128/vshow}{GSS Data Explorer page for \code{intprfssnl}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab almost always \tab often \tab sometimes \tab hardly ever \tab never \tab no access \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 587 \tab 252 \tab 159 \tab 80 \tab 323 \tab 56 \tab 93 \tab 1731 \tab 10 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 587 \tab 252 \tab 159 \tab 80 \tab 323 \tab 56 \tab 93 \tab 1731 \tab 10 \tab 18 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{inthome}},
\code{\link{intliteracy}},
\code{\link{intpub}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplgrsn}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
