% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intpub}
\alias{intpub}
\title{Frequency r has internet access in public}
\source{
General Social Survey https://gss.norc.org
}
\description{
intpub
}
\details{
Question How often, if at all, do you typically use the Internet on any device at the following locations for any reason?… In public places?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{6} no access
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8129/vshow}{GSS Data Explorer page for \code{intpub}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{intcomm}},
\code{\link{intliteracy}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplgrsn}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
