% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_602.R
\name{intrace1}
\alias{intrace1}
\title{Interviewer s race1}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrace1
}
\details{
Question What is your race?  Indicate one or more races that you consider yourself to be.
A. First mention.
}
\section{Values}{

\itemize{
\item \verb{[1]} white
\item \verb{[2]} black or african-american
\item \verb{[3]} american indian or alaska native
\item \verb{[4]} asian indian
\item \verb{[5]} chinese
\item \verb{[6]} filipino
\item \verb{[7]} japanese
\item \verb{[8]} korean
\item \verb{[9]} vietnamese
\item \verb{[10]} other asian
\item \verb{[11]} native hawaiian
\item \verb{[12]} guamanian or chamorro
\item \verb{[13]} samoan
\item \verb{[14]} other pacific islander
\item \verb{[15]} some other race
\item \verb{[16]} hispanic
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5512/vshow}{GSS Data Explorer page for \code{intrace1}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab white \tab black or african-american \tab american indian or alaska native \tab asian indian \tab chinese \tab filipino \tab vietnamese \tab other asian \tab some other race \tab hispanic \tab don't know \tab no answer \tab japanese \tab other pacific islander \tab korean \tab not available in this release \tab Total \cr
   2004 \tab - \tab 2382 \tab 377 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 23 \tab 1 \tab 18 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 3675 \tab 556 \tab 60 \tab 38 \tab 8 \tab 1 \tab 14 \tab 3 \tab 70 \tab 66 \tab - \tab 18 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1707 \tab 269 \tab 1 \tab 19 \tab - \tab 7 \tab 1 \tab - \tab 7 \tab 9 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1618 \tab 282 \tab 27 \tab 23 \tab 16 \tab 7 \tab - \tab 2 \tab 26 \tab 10 \tab 1 \tab 31 \tab - \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1611 \tab 232 \tab 53 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 25 \tab 4 \tab 1 \tab 37 \tab 5 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2152 \tab 259 \tab 13 \tab 10 \tab 1 \tab - \tab - \tab 6 \tab 25 \tab 16 \tab 1 \tab 34 \tab - \tab - \tab 21 \tab - \tab 2538 \cr
   2016 \tab - \tab 2162 \tab 458 \tab 4 \tab 128 \tab 4 \tab 2 \tab 2 \tab 2 \tab 34 \tab 19 \tab 37 \tab 14 \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 1845 \tab 347 \tab 1 \tab 51 \tab 9 \tab 7 \tab - \tab 1 \tab 41 \tab - \tab 2 \tab 43 \tab - \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 0 \tab 17152 \tab 2780 \tab 160 \tab 276 \tab 41 \tab 25 \tab 19 \tab 15 \tab 230 \tab 147 \tab 43 \tab 198 \tab 7 \tab 2 \tab 21 \tab 4032 \tab 25148 \cr
}
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp1a}},
\code{\link{bthgrp1b}},
\code{\link{bthgrp1c}},
\code{\link{bthgrp1d}},
\code{\link{bthgrp1e}},
\code{\link{bthgrp1f}},
\code{\link{bthgrp2a}},
\code{\link{intrace2}},
\code{\link{intrace3}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
