% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_602.R
\name{intrace3}
\alias{intrace3}
\title{Interviewer s race3}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrace3
}
\details{
Question What is your race?  Indicate one or more races that you consider yourself to be.
C. Third mention.
}
\section{Values}{

\itemize{
\item \verb{[1]} white
\item \verb{[2]} black or african-american
\item \verb{[3]} american indian or alaska native
\item \verb{[4]} asian indian
\item \verb{[5]} chinese
\item \verb{[6]} filipino
\item \verb{[7]} japanese
\item \verb{[8]} korean
\item \verb{[9]} vietnamese
\item \verb{[10]} other asian
\item \verb{[11]} native hawaiian
\item \verb{[12]} guamanian or chamorro
\item \verb{[13]} samoan
\item \verb{[14]} other pacific islander
\item \verb{[15]} some other race
\item \verb{[16]} hispanic
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5514/vshow}{GSS Data Explorer page for \code{intrace3}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab white \tab black or african-american \tab american indian or alaska native \tab hispanic \tab no answer \tab some other race \tab don't know \tab asian indian \tab not available in this release \tab Total \cr
   2004 \tab 2789 \tab 1 \tab 2 \tab 1 \tab 1 \tab 18 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4485 \tab 1 \tab 1 \tab 2 \tab 3 \tab 18 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2043 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1973 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2535 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 2851 \tab - \tab - \tab 1 \tab - \tab 14 \tab - \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 2296 \tab 1 \tab - \tab 3 \tab - \tab 43 \tab 2 \tab 2 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 20995 \tab 3 \tab 4 \tab 8 \tab 4 \tab 93 \tab 4 \tab 4 \tab 1 \tab 4032 \tab 25148 \cr
}
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp1a}},
\code{\link{bthgrp1b}},
\code{\link{bthgrp1c}},
\code{\link{bthgrp1d}},
\code{\link{bthgrp1e}},
\code{\link{bthgrp1f}},
\code{\link{bthgrp2a}},
\code{\link{intrace1}},
\code{\link{intrace2}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
