% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_286.R
\name{intrecnt}
\alias{intrecnt}
\title{Used internet/apps yesterday}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrecnt
}
\details{
Question Did you happen to use the internet or web-enabled applications/APPS yesterday?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, used internet yesterday
\item \verb{[2]} no, did not use internet yesterday
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6268/vshow}{GSS Data Explorer page for \code{intrecnt}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes, used internet yesterday \tab no, did not use internet yesterday \tab no answer \tab Total \cr
   2016 \tab 1483 \tab 1261 \tab 114 \tab 9 \tab 2867 \cr
   Total \tab 1483 \tab 1261 \tab 114 \tab 9 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{facebook}},
\code{\link{googlesn}},
\code{\link{instagrm}},
\code{\link{linkedin}},
\code{\link{snapchat}},
\code{\link{tumblr}},
\code{\link{twitter}},
\code{\link{whatsapp}}
}
\concept{Social Media Usage}
\keyword{variable}
