% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{intrhome}
\alias{intrhome}
\title{Internet access in r's home}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrhome
}
\details{
Question Now for a few items on you. Do you have access to the Internet in your home?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3346/vshow}{GSS Data Explorer page for \code{intrhome}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1187 \tab 675 \tab 1 \tab 1 \tab 4510 \cr
   2008 \tab 518 \tab 1031 \tab 461 \tab 6 \tab 7 \tab 2023 \cr
   2010 \tab 1328 \tab 527 \tab 185 \tab 1 \tab 3 \tab 2044 \cr
   2012 \tab 972 \tab 756 \tab 237 \tab 4 \tab 5 \tab 1974 \cr
   2014 \tab 1299 \tab 1011 \tab 227 \tab - \tab 1 \tab 2538 \cr
   Total \tab 6763 \tab 4512 \tab 1785 \tab 12 \tab 17 \tab 13089 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
