% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_675.R
\name{intrust}
\alias{intrust}
\title{How much r trusts people they've only met online}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrust
}
\details{
Question On a scale of 0 to 10, how much do you trust people you are communicating with on the Internet but have never met in person? 0 means you do not trust them at all, and 10 means you trust them completely.
}
\section{Values}{

\itemize{
\item \code{0} no trust at all
\item \code{10} complete trust
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8175/vshow}{GSS Data Explorer page for \code{intrust}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intcnfrm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{intlnly}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intsearch}},
\code{\link{intstream}},
\code{\link{inttrig}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
