% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{intrwght}
\alias{intrwght}
\title{Interview assessment of r's weight}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrwght
}
\details{
Question Rate respondents weight without asking: IF R IS PREGNANT CONSIDER HER WEIGHT ASIDE FROM BEING PREGNANT.
}
\section{Values}{

\itemize{
\item \verb{[1]} below average
\item \verb{[2]} average
\item \verb{[3]} somewhat above average
\item \verb{[4]} considerably above average
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3052/vshow}{GSS Data Explorer page for \code{intrwght}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab below average \tab average \tab somewhat above average \tab considerably above average \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2004 \tab 432 \tab 156 \tab 1556 \tab 457 \tab 96 \tab 113 \tab 2 \tab - \tab 2812 \cr
   Total \tab 432 \tab 156 \tab 1556 \tab 457 \tab 96 \tab 113 \tab 2 \tab 0 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{genevig1}},
\code{\link{genevig2}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{owndoing}},
\code{\link{pessimst}},
\code{\link{sadblue}},
\code{\link{satself}},
\code{\link{selfish}},
\code{\link{slfrspct}}

Other Interview: 
\code{\link{comprend}},
\code{\link{consent}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{id}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{inthisp}},
\code{\link{intid}},
\code{\link{intrace1}},
\code{\link{intrace2}},
\code{\link{intrace3}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{intyrs_0418}},
\code{\link{lngthinv}},
\code{\link{mode}()},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{sciintro}},
\code{\link{scimode}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{year}()}
}
\concept{Genes and the Environment}
\concept{Interview}
\keyword{variable}
