% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{intscam}
\alias{intscam}
\title{Has r or someone they know been a victim of online scam}
\source{
General Social Survey https://gss.norc.org
}
\description{
intscam
}
\details{
Question Some people receive emails or text messages from scammers from time to time, some of them also become victims. Apart from only receiving such messages, have you or someone you know personally ever been a victim of an online scam, identity theft or hacking attack?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, me myself
\item \verb{[2]} yes, someone i know personally
\item \verb{[3]} yes, me and someone else i know personally
\item \verb{[4]} no
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8126/vshow}{GSS Data Explorer page for \code{intscam}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab yes, me myself \tab yes, someone i know personally \tab yes, me and someone else i know personally \tab no \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 315 \tab 579 \tab 148 \tab 508 \tab 76 \tab 1670 \tab 5 \tab 8 \tab 3309 \cr
   Total \tab 0 \tab 315 \tab 579 \tab 148 \tab 508 \tab 76 \tab 1670 \tab 5 \tab 8 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{aidrive}},
\code{\link{aimed}},
\code{\link{aiworry}},
\code{\link{classtech}},
\code{\link{dataprot}},
\code{\link{eductech}},
\code{\link{emonitor1}},
\code{\link{ftruse}},
\code{\link{gendtech}},
\code{\link{harmgood1}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{intaccprxy}},
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{intcnfrm}},
\code{\link{intcomm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{inthome}},
\code{\link{inthrss}},
\code{\link{intliteracy}},
\code{\link{intlnly}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intrnetuse}},
\code{\link{intrust}},
\code{\link{intsearch}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{intstream}},
\code{\link{intsurf}},
\code{\link{inttrig}},
\code{\link{intviews}},
\code{\link{modelobby}},
\code{\link{modeorg}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{nextgen1}},
\code{\link{nouseint1}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint2}},
\code{\link{nouseint3}},
\code{\link{nouseint4}},
\code{\link{nouseint5}},
\code{\link{nouseint6}},
\code{\link{nouseint7}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{papernews}},
\code{\link{polnewsfrom}},
\code{\link{prxyuse}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{techesy}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{tvnews1}},
\code{\link{unplgrsn}},
\code{\link{unplug}},
\code{\link{urbantech}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
