% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{intsci}
\alias{intsci}
\title{Interested in new scientific discoveries}
\source{
General Social Survey https://gss.norc.org
}
\description{
intsci
}
\details{
Question Issues about new scientific discoveries. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} moderately interested
\item \verb{[3]} not at all interested
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3455/vshow}{GSS Data Explorer page for \code{intsci}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 518 \tab 607 \tab 675 \tab 213 \tab 9 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1334 \tab 273 \tab 367 \tab 62 \tab 2 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 409 \tab 431 \tab 148 \tab 9 \tab 5 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 495 \tab 567 \tab 175 \tab 1 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 585 \tab 583 \tab 217 \tab 2 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 476 \tab 520 \tab 177 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab 2144 \tab 849 \tab 858 \tab 170 \tab - \tab - \tab 11 \tab 4032 \cr
   Total \tab 8917 \tab 3694 \tab 4001 \tab 1162 \tab 25 \tab 16 \tab 11 \tab 17826 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{scimode}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
