% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_675.R
\name{intsearch}
\alias{intsearch}
\title{How often r uses internet for search}
\source{
General Social Survey https://gss.norc.org
}
\description{
intsearch
}
\details{
Question How would you rate your ability to do the following?… Searching for information?
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8119/vshow}{GSS Data Explorer page for \code{intsearch}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{aidrive}},
\code{\link{aimed}},
\code{\link{aiworry}},
\code{\link{classtech}},
\code{\link{dataprot}},
\code{\link{eductech}},
\code{\link{emonitor1}},
\code{\link{ftruse}},
\code{\link{gendtech}},
\code{\link{harmgood1}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{intaccprxy}},
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{intcnfrm}},
\code{\link{intcomm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{inthome}},
\code{\link{inthrss}},
\code{\link{intliteracy}},
\code{\link{intlnly}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intrnetuse}},
\code{\link{intrust}},
\code{\link{intscam}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{intstream}},
\code{\link{intsurf}},
\code{\link{inttrig}},
\code{\link{intviews}},
\code{\link{modelobby}},
\code{\link{modeorg}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{nextgen1}},
\code{\link{nouseint1}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint2}},
\code{\link{nouseint3}},
\code{\link{nouseint4}},
\code{\link{nouseint5}},
\code{\link{nouseint6}},
\code{\link{nouseint7}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{papernews}},
\code{\link{polnewsfrom}},
\code{\link{prxyuse}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{techesy}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{tvnews1}},
\code{\link{unplgrsn}},
\code{\link{unplug}},
\code{\link{urbantech}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
