% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{intsex}
\alias{intsex}
\title{Sex of interviewer}
\source{
General Social Survey https://gss.norc.org
}
\description{
intsex
}
\details{
Question C. Sex of interviewer
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5567/vshow}{GSS Data Explorer page for \code{intsex}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab not available in this release \tab Total \cr
   2004 \tab - \tab 471 \tab 2271 \tab 70 \tab - \tab 2812 \cr
   2006 \tab - \tab 875 \tab 3348 \tab 287 \tab - \tab 4510 \cr
   2008 \tab - \tab 305 \tab 1714 \tab 4 \tab - \tab 2023 \cr
   2010 \tab - \tab 398 \tab 1619 \tab 27 \tab - \tab 2044 \cr
   2012 \tab - \tab 323 \tab 1651 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 557 \tab 1981 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 319 \tab 2548 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 318 \tab 2030 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 0 \tab 3566 \tab 17162 \tab 388 \tab 4032 \tab 25148 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intyrs}},
\code{\link{lngthinv}},
\code{\link{mode}()}
}
\concept{Paradata}
\keyword{variable}
