% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intskill}
\alias{intskill}
\title{R's skill at using the internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
intskill
}
\details{
Question How would you rate your ability to do the following?… Using the internet?
}
\section{Values}{

\itemize{
\item \verb{[1]} very good
\item \verb{[2]} good
\item \verb{[3]} neither good nor bad
\item \verb{[4]} bad
\item \verb{[5]} very bad
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8149/vshow}{GSS Data Explorer page for \code{intskill}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab very good \tab good \tab neither good nor bad \tab bad \tab very bad \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 832 \tab 442 \tab 210 \tab 39 \tab 21 \tab 23 \tab 1732 \tab 4 \tab 6 \tab 3309 \cr
   Total \tab 0 \tab 832 \tab 442 \tab 210 \tab 39 \tab 21 \tab 23 \tab 1732 \tab 4 \tab 6 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{inthome}},
\code{\link{intliteracy}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intsurf}},
\code{\link{unplgrsn}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
