% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_382.R
\name{intspace}
\alias{intspace}
\title{Interested in space exploration}
\source{
General Social Survey https://gss.norc.org
}
\description{
intspace
}
\details{
Question Issues about space exploration. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} moderately interested
\item \verb{[3]} not at all interested
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3459/vshow}{GSS Data Explorer page for \code{intspace}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 326 \tab 665 \tab 508 \tab 5 \tab 1 \tab 2023 \cr
   2010 \tab 1334 \tab 148 \tab 365 \tab 189 \tab 2 \tab 6 \tab 2044 \cr
   2012 \tab 972 \tab 221 \tab 433 \tab 332 \tab 11 \tab 5 \tab 1974 \cr
   2014 \tab 1299 \tab 272 \tab 559 \tab 407 \tab - \tab 1 \tab 2538 \cr
   2016 \tab 1477 \tab 342 \tab 600 \tab 443 \tab 2 \tab 3 \tab 2867 \cr
   2018 \tab 1173 \tab 300 \tab 524 \tab 349 \tab 2 \tab - \tab 2348 \cr
   Total \tab 6773 \tab 1609 \tab 3146 \tab 2228 \tab 22 \tab 16 \tab 13794 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intenvir}},
\code{\link{intmil}},
\code{\link{sciintro}},
\code{\link{scimath}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
