% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intsurf}
\alias{intsurf}
\title{R's skill at online searches}
\source{
General Social Survey https://gss.norc.org
}
\description{
intsurf
}
\details{
Question How would you rate your ability to do the following?… Doing online searches?
}
\section{Values}{

\itemize{
\item \code{1} very good
\item \code{2} good
\item \code{3} neither good nor bad
\item \code{4} bad
\item \code{5} very bad
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8150/vshow}{GSS Data Explorer page for \code{intsurf}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{intcomm}},
\code{\link{intliteracy}},
\code{\link{intpub}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{unplgrsn}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
