% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{inttech}
\alias{inttech}
\title{Interested in technologies}
\source{
General Social Survey https://gss.norc.org
}
\description{
inttech
}
\details{
Question Issues about the use of new inventions and technologies. (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} moderately interested
\item \verb{[3]} not at all interested
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3457/vshow}{GSS Data Explorer page for \code{inttech}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 518 \tab 626 \tab 680 \tab 193 \tab 5 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1334 \tab 270 \tab 367 \tab 66 \tab 1 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 420 \tab 454 \tab 113 \tab 11 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 530 \tab 559 \tab 149 \tab - \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 570 \tab 637 \tab 179 \tab 1 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 468 \tab 549 \tab 156 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab 2146 \tab 782 \tab 915 \tab 177 \tab - \tab - \tab 12 \tab 4032 \cr
   Total \tab 8919 \tab 3666 \tab 4161 \tab 1033 \tab 20 \tab 15 \tab 12 \tab 17826 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{scimode}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
