% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_285.R
\name{intuse}
\alias{intuse}
\title{Use internet/apps more than occassionally}
\source{
General Social Survey https://gss.norc.org
}
\description{
intuse
}
\details{
Question Do you use the internet or web-enabled applications, known as APPS, for more than email, at least occasionally?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6266/vshow}{GSS Data Explorer page for \code{intuse}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab 911 \tab 1376 \tab 570 \tab 2 \tab 8 \tab - \tab 2867 \cr
   Total \tab 911 \tab 1376 \tab 570 \tab 2 \tab 8 \tab 0 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{clssmtes}},
\code{\link{facebook}},
\code{\link{flickr}},
\code{\link{googlesn}},
\code{\link{instagrm}},
\code{\link{intmbile}},
\code{\link{intrecnt}},
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{intwkenm}},
\code{\link{linkedin}},
\code{\link{pinterst}},
\code{\link{snapchat}},
\code{\link{snsmfrst}},
\code{\link{snsmnew}},
\code{\link{snsmot2a}},
\code{\link{snsmot2b}},
\code{\link{snsmot2c}},
\code{\link{snsmoth1}},
\code{\link{tumblr}},
\code{\link{twitter}},
\code{\link{vine}},
\code{\link{whatsapp}}
}
\concept{Social Media Usage}
\keyword{variable}
