% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{intviews}
\alias{intviews}
\title{How often r talks about politics online}
\source{
General Social Survey https://gss.norc.org
}
\description{
intviews
}
\details{
Question How often, if at all, did you express, support or oppose political views on the Internet or social media in the last 12 months?
}
\section{Values}{

\itemize{
\item \code{1} every day or almost every day
\item \code{2} several times a week
\item \code{3} several times a month
\item \code{4} several times a year
\item \code{5} less often
\item \code{6} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8177/vshow}{GSS Data Explorer page for \code{intviews}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{inthrss}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intscam}},
\code{\link{nextgen1}},
\code{\link{polnewsfrom}},
\code{\link{smnews}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
