% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{intwkenm}
\alias{intwkenm}
\title{Minutes of internet use on weekends}
\source{
General Social Survey https://gss.norc.org
}
\description{
intwkenm
}
\details{
Question MINUTES:
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6289/vshow}{GSS Data Explorer page for \code{intwkenm}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 15 \tab 20 \tab 25 \tab 3 \tab 30 \tab 35 \tab 40 \tab 45 \tab 5 \tab 50 \tab 59 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1630 \tab 1018 \tab 1 \tab 15 \tab 14 \tab 16 \tab 2 \tab 2 \tab 134 \tab 1 \tab 3 \tab 11 \tab 6 \tab 1 \tab 1 \tab 4 \tab 8 \tab 2867 \cr
   Total \tab 1630 \tab 1018 \tab 1 \tab 15 \tab 14 \tab 16 \tab 2 \tab 2 \tab 134 \tab 1 \tab 3 \tab 11 \tab 6 \tab 1 \tab 1 \tab 4 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{snsmfrst}}
}
\concept{Social Media Usage}
\keyword{variable}
