% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_695.R
\name{intyrs_0418}
\alias{intyrs_0418}
\title{Years of service as an interviewer at norc}
\source{
General Social Survey https://gss.norc.org
}
\description{
intyrs_0418
}
\details{
Question D. Years of service as an interviewer at NORC
As of 2024, INTYRS has been recoded to only account for fewer categories. This has been retroactively applied to all previous years of INTYRS. The original INTYRS containing data from 2004 to 2018 has been renamed INTYRS_0418.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} i don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8286/vshow}{GSS Data Explorer page for \code{intyrs_0418}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 0.1 \tab 0.2 \tab 0.3 \tab 0.4 \tab 0.5 \tab 0.6 \tab 0.7 \tab 0.9 \tab 1 \tab 1.1 \tab 1.2 \tab 1.3 \tab 1.6 \tab 1.9 \tab 11.5 \tab 11.8 \tab 12 \tab 2.1 \tab 2.4 \tab 2.7 \tab 2.9 \tab 3 \tab 3.1 \tab 3.4 \tab 3.7 \tab 4.3 \tab 4.8 \tab 5.2 \tab 6.8 \tab 6.9 \tab 7.5 \tab 7.6 \tab 7.9 \tab 8.1 \tab 8.9 \tab 9.3 \tab 9.5 \tab 9.6 \tab 9.8 \tab no answer \tab 0.8 \tab 1.8 \tab 10.7 \tab 11.1 \tab 11.2 \tab 11.3 \tab 11.7 \tab 11.9 \tab 13.1 \tab 13.6 \tab 17.2 \tab 17.8 \tab 19.6 \tab 2 \tab 2.2 \tab 2.5 \tab 2.6 \tab 2.8 \tab 3.2 \tab 3.5 \tab 3.9 \tab 4.2 \tab 4.4 \tab 4.5 \tab 4.6 \tab 4.7 \tab 5 \tab 5.1 \tab 5.3 \tab 5.8 \tab 6.3 \tab 6.5 \tab 7 \tab 8.2 \tab 8.4 \tab 8.5 \tab 9.1 \tab 9.7 \tab 0.12 \tab 0.16 \tab 0.23 \tab 0.26 \tab 0.27 \tab 0.31 \tab 0.35 \tab 0.38 \tab 0.39 \tab 0.42 \tab 0.43 \tab 0.44 \tab 0.61 \tab 0.97 \tab 1.03 \tab 1.06 \tab 1.11 \tab 1.39 \tab 1.92 \tab 10.29 \tab 11.41 \tab 11.74 \tab 13.22 \tab 13.95 \tab 14.97 \tab 15.52 \tab 15.71 \tab 18.72 \tab 19.87 \tab 2.01 \tab 2.03 \tab 2.36 \tab 2.38 \tab 2.54 \tab 2.94 \tab 2.99 \tab 25.64 \tab 3.06 \tab 3.09 \tab 3.16 \tab 3.67 \tab 3.78 \tab 3.82 \tab 3.87 \tab 4.02 \tab 4.05 \tab 4.1 \tab 4.17 \tab 4.29 \tab 4.35 \tab 4.62 \tab 4.65 \tab 5.09 \tab 5.26 \tab 5.28 \tab 5.29 \tab 5.6 \tab 5.88 \tab 6.27 \tab 6.39 \tab 6.54 \tab 6.63 \tab 6.67 \tab 6.68 \tab 6.72 \tab 6.81 \tab 6.89 \tab 7.07 \tab 7.11 \tab 7.39 \tab 7.87 \tab 8.62 \tab 8.67 \tab 8.87 \tab 8.93 \tab 0.517808219178082 \tab 0.520547945205479 \tab 0.523287671232877 \tab 0.83013698630137 \tab 0.857534246575342 \tab 0.906849315068493 \tab 0.986301369863014 \tab 1.04931506849315 \tab 1.1041095890411 \tab 1.11780821917808 \tab 1.12602739726027 \tab 1.16164383561644 \tab 1.19452054794521 \tab 1.44383561643836 \tab 1.5 \tab 1.96259124087591 \tab 4.87448653582839 \tab 5.09356458238247 \tab 8.41131731061758 \tab 9 \tab 10 \tab 11 \tab 14 \tab 15 \tab 17 \tab 19 \tab 23 \tab 4 \tab 6 \tab 8 \tab 13 \tab 16 \tab 25 \tab 20 \tab 21 \tab Total \cr
   2004 \tab - \tab 5 \tab 73 \tab 142 \tab 131 \tab 1005 \tab 18 \tab 127 \tab 66 \tab 66 \tab 3 \tab 11 \tab 101 \tab 3 \tab 16 \tab 6 \tab 31 \tab 16 \tab 70 \tab 13 \tab 12 \tab 24 \tab 79 \tab 114 \tab 85 \tab 40 \tab 34 \tab 7 \tab 19 \tab 33 \tab 26 \tab 45 \tab 35 \tab 31 \tab 19 \tab 58 \tab 5 \tab 15 \tab 44 \tab 64 \tab 50 \tab 70 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab 21 \tab 784 \tab - \tab - \tab - \tab 136 \tab 258 \tab 127 \tab - \tab - \tab - \tab - \tab 11 \tab 72 \tab - \tab - \tab - \tab 5 \tab - \tab 34 \tab - \tab - \tab 41 \tab 12 \tab 34 \tab 18 \tab 89 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 26 \tab - \tab - \tab 10 \tab 167 \tab 15 \tab 55 \tab 21 \tab 90 \tab 9 \tab 5 \tab 46 \tab 76 \tab 42 \tab 1 \tab 12 \tab 653 \tab 451 \tab 40 \tab 64 \tab 24 \tab 90 \tab 138 \tab 11 \tab 3 \tab 60 \tab 106 \tab 36 \tab 37 \tab 13 \tab 75 \tab 121 \tab 51 \tab 15 \tab 67 \tab 17 \tab 48 \tab 39 \tab 2 \tab 44 \tab 81 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab 25 \tab - \tab 24 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 16 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 23 \tab 24 \tab 95 \tab 9 \tab 60 \tab 10 \tab 33 \tab 10 \tab 86 \tab 11 \tab 155 \tab 244 \tab 14 \tab 1 \tab 1 \tab 2 \tab 2 \tab 50 \tab 24 \tab 42 \tab 2 \tab 17 \tab 23 \tab 42 \tab 13 \tab 23 \tab 15 \tab 12 \tab 19 \tab 6 \tab 24 \tab 8 \tab 21 \tab 1 \tab 14 \tab 22 \tab 10 \tab 35 \tab 14 \tab 33 \tab 8 \tab 3 \tab 8 \tab 34 \tab 2 \tab 59 \tab 117 \tab 14 \tab 48 \tab 56 \tab 18 \tab 10 \tab 18 \tab 11 \tab 16 \tab 22 \tab 20 \tab 5 \tab 13 \tab 3 \tab 16 \tab 33 \tab 14 \tab 1 \tab 1 \tab 13 \tab 9 \tab 13 \tab 18 \tab 17 \tab 15 \tab 11 \tab 30 \tab 13 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab 13 \tab - \tab - \tab - \tab 161 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 27 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 13 \tab 35 \tab 1279 \tab 7 \tab 14 \tab 28 \tab 29 \tab 39 \tab 5 \tab 21 \tab 34 \tab 199 \tab 3 \tab 4 \tab 3 \tab 12 \tab 32 \tab 10 \tab 13 \tab 39 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 560 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 117 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 28 \tab - \tab - \tab - \tab - \tab 106 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 399 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 23 \tab - \tab - \tab - \tab - \tab - \tab 79 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 25 \tab 61 \tab 51 \tab 18 \tab 3 \tab 30 \tab 21 \tab 13 \tab 139 \tab 96 \tab 205 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 340 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 123 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 67 \tab - \tab - \tab - \tab - \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 501 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 117 \tab - \tab - \tab - \tab - \tab - \tab 219 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 103 \tab 151 \tab 41 \tab - \tab - \tab 33 \tab 7 \tab - \tab 523 \tab 185 \tab 47 \tab 25 \tab 25 \tab 16 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 751 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 166 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 122 \tab - \tab - \tab - \tab - \tab 77 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 336 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 40 \tab - \tab - \tab - \tab - \tab - \tab 123 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 209 \tab 104 \tab - \tab 10 \tab - \tab - \tab 14 \tab 22 \tab 189 \tab 346 \tab 205 \tab 79 \tab 19 \tab - \tab 55 \tab - \tab 2867 \cr
   2018 \tab - \tab 680 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 33 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 31 \tab - \tab - \tab - \tab - \tab 27 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 542 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 94 \tab - \tab - \tab - \tab - \tab - \tab 33 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 48 \tab 21 \tab 62 \tab 139 \tab 34 \tab - \tab - \tab 30 \tab 179 \tab 165 \tab 136 \tab 2 \tab - \tab 8 \tab 4 \tab 80 \tab 2348 \cr
   Total \tab 0 \tab 2336 \tab 73 \tab 167 \tab 152 \tab 1813 \tab 31 \tab 127 \tab 66 \tab 202 \tab 861 \tab 138 \tab 101 \tab 3 \tab 16 \tab 6 \tab 42 \tab 88 \tab 318 \tab 13 \tab 12 \tab 29 \tab 79 \tab 373 \tab 85 \tab 40 \tab 75 \tab 19 \tab 53 \tab 51 \tab 115 \tab 52 \tab 35 \tab 31 \tab 19 \tab 58 \tab 5 \tab 15 \tab 44 \tab 90 \tab 50 \tab 101 \tab 10 \tab 167 \tab 15 \tab 55 \tab 21 \tab 90 \tab 9 \tab 5 \tab 46 \tab 92 \tab 42 \tab 1 \tab 12 \tab 2433 \tab 451 \tab 40 \tab 64 \tab 24 \tab 90 \tab 138 \tab 11 \tab 3 \tab 60 \tab 110 \tab 36 \tab 37 \tab 287 \tab 75 \tab 121 \tab 51 \tab 15 \tab 67 \tab 489 \tab 48 \tab 39 \tab 2 \tab 44 \tab 81 \tab 23 \tab 24 \tab 95 \tab 9 \tab 60 \tab 10 \tab 33 \tab 10 \tab 86 \tab 11 \tab 155 \tab 244 \tab 14 \tab 1 \tab 1 \tab 2 \tab 2 \tab 50 \tab 24 \tab 42 \tab 2 \tab 17 \tab 23 \tab 42 \tab 13 \tab 23 \tab 15 \tab 12 \tab 19 \tab 6 \tab 24 \tab 8 \tab 21 \tab 1 \tab 14 \tab 22 \tab 10 \tab 35 \tab 14 \tab 33 \tab 8 \tab 3 \tab 8 \tab 34 \tab 2 \tab 59 \tab 117 \tab 14 \tab 48 \tab 56 \tab 18 \tab 10 \tab 18 \tab 11 \tab 16 \tab 22 \tab 20 \tab 5 \tab 13 \tab 3 \tab 16 \tab 33 \tab 14 \tab 1 \tab 1 \tab 13 \tab 9 \tab 13 \tab 18 \tab 17 \tab 15 \tab 11 \tab 30 \tab 13 \tab 15 \tab 13 \tab 35 \tab 1279 \tab 7 \tab 14 \tab 28 \tab 29 \tab 39 \tab 5 \tab 21 \tab 34 \tab 199 \tab 3 \tab 4 \tab 3 \tab 12 \tab 32 \tab 10 \tab 13 \tab 424 \tab 337 \tab 154 \tab 167 \tab 37 \tab 63 \tab 42 \tab 65 \tab 1030 \tab 792 \tab 593 \tab 106 \tab 44 \tab 24 \tab 59 \tab 80 \tab 21116 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{devtype}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{huclean}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{intid}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}},
\code{\link{mode}()},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse1}},
\code{\link{whoelse2}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}

Other Interview: 
\code{\link{comprend}},
\code{\link{consent}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{id}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{inthisp}},
\code{\link{intid}},
\code{\link{intrace1}},
\code{\link{intrace2}},
\code{\link{intrace3}},
\code{\link{intrwght}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}},
\code{\link{mode}()},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{sciintro}},
\code{\link{scimode}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{year}()}
}
\concept{Interview}
\concept{Paradata}
\keyword{variable}
