% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_266.R
\name{invest12}
\alias{invest12}
\title{R use web to invest money in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
invest12
}
\details{
Question In the past 12 months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/02 have you used the Web to do each of these things:
E. Invest money in a stock or mutual fund?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2419/vshow}{GSS Data Explorer page for \code{invest12}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 113 \tab 553 \tab 76 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 129 \tab 860 \tab 94 \tab 1 \tab 2765 \cr
   Total \tab 3756 \tab 242 \tab 1413 \tab 170 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{person30}},
\code{\link{porn30}},
\code{\link{sci30}},
\code{\link{work12}}
}
\concept{Information Society}
\keyword{variable}
