% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{irrelart}
\alias{irrelart}
\title{Should allow art to mock or demean religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
irrelart
}
\details{
Question Consider an artist--like a painter, sculptor, or photographer--whose art works mock or demean religions or religious figures. If such a person wanted to exhibit these works in your community, should the exhibit be allowed, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} allowed
\item \verb{[2]} not allowed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2126/vshow}{GSS Data Explorer page for \code{irrelart}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab allowed \tab not allowed \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 779 \tab 638 \tab 15 \tab 13 \tab 2832 \cr
   Total \tab 1387 \tab 779 \tab 638 \tab 15 \tab 13 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{irreloff}},
\code{\link{natarts}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion and Culture}
\keyword{variable}
