% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{irreloff}
\alias{irreloff}
\title{Offended art mocks or demeans religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
irreloff
}
\details{
Question Consider an artist--like a painter, sculptor, or photographer--whose art works mock or demean religions or religious figures. If such a person wanted to exhibit these works in your community, should the exhibit be allowed, or not?
A. Would you, personally, be offended by an exhibit of such artworks or would you not be offended?
}
\section{Values}{

\itemize{
\item \verb{[1]} offended
\item \verb{[2]} not offended
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2127/vshow}{GSS Data Explorer page for \code{irreloff}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab offended \tab not offended \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2040 \tab 316 \tab 444 \tab 10 \tab 22 \tab 2832 \cr
   Total \tab 2040 \tab 316 \tab 444 \tab 10 \tab 22 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{irrelart}},
\code{\link{natarts}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion and Culture}
\keyword{variable}
