% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{israel}
\alias{israel}
\title{Liking for israel}
\source{
General Social Survey https://gss.norc.org
}
\description{
israel
}
\details{
Question You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? READ EACH ITEM.
G. Israel
}
\section{Values}{

\itemize{
\item \verb{[0]} +5
\item \verb{[1]} +4
\item \verb{[2]} +3
\item \verb{[3]} +2
\item \verb{[4]} +1
\item \verb{[5]} -1
\item \verb{[6]} -2
\item \verb{[7]} -3
\item \verb{[8]} -4
\item \verb{[9]} -5
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1974, 1975, 1977, 1982, 1983, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5496/vshow}{GSS Data Explorer page for \code{israel}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab +5 \tab +4 \tab +3 \tab +2 \tab +1 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab Total \cr
   1974 \tab - \tab 198 \tab 141 \tab 231 \tab 175 \tab 258 \tab 128 \tab 57 \tab 67 \tab 33 \tab 89 \tab 107 \tab 1484 \cr
   1975 \tab - \tab 146 \tab 123 \tab 205 \tab 169 \tab 275 \tab 164 \tab 82 \tab 77 \tab 48 \tab 76 \tab 125 \tab 1490 \cr
   1977 \tab - \tab 180 \tab 151 \tab 204 \tab 173 \tab 270 \tab 149 \tab 71 \tab 64 \tab 41 \tab 69 \tab 158 \tab 1530 \cr
   1982 \tab - \tab 207 \tab 147 \tab 250 \tab 211 \tab 290 \tab 215 \tab 121 \tab 92 \tab 62 \tab 95 \tab 170 \tab 1860 \cr
   1983 \tab - \tab 148 \tab 123 \tab 169 \tab 172 \tab 285 \tab 209 \tab 106 \tab 111 \tab 67 \tab 103 \tab 106 \tab 1599 \cr
   1985 \tab - \tab 168 \tab 125 \tab 211 \tab 201 \tab 254 \tab 160 \tab 94 \tab 82 \tab 48 \tab 75 \tab 116 \tab 1534 \cr
   1986 \tab - \tab 165 \tab 128 \tab 175 \tab 180 \tab 285 \tab 144 \tab 84 \tab 73 \tab 53 \tab 104 \tab 79 \tab 1470 \cr
   1988 \tab - \tab 96 \tab 60 \tab 101 \tab 91 \tab 155 \tab 130 \tab 77 \tab 77 \tab 30 \tab 94 \tab 570 \tab 1481 \cr
   1989 \tab - \tab 87 \tab 56 \tab 107 \tab 95 \tab 177 \tab 104 \tab 70 \tab 59 \tab 43 \tab 118 \tab 621 \tab 1537 \cr
   1990 \tab - \tab 75 \tab 48 \tab 86 \tab 93 \tab 182 \tab 128 \tab 69 \tab 45 \tab 30 \tab 93 \tab 523 \tab 1372 \cr
   1991 \tab - \tab 152 \tab 104 \tab 162 \tab 138 \tab 177 \tab 76 \tab 38 \tab 35 \tab 14 \tab 55 \tab 566 \tab 1517 \cr
   1993 \tab - \tab 105 \tab 73 \tab 133 \tab 104 \tab 227 \tab 123 \tab 71 \tab 59 \tab 41 \tab 66 \tab 604 \tab 1606 \cr
   1994 \tab - \tab 46 \tab 32 \tab 52 \tab 59 \tab 110 \tab 55 \tab 31 \tab 34 \tab 14 \tab 34 \tab 2525 \tab 2992 \cr
   Total \tab 0 \tab 1773 \tab 1311 \tab 2086 \tab 1861 \tab 2945 \tab 1785 \tab 971 \tab 875 \tab 524 \tab 1071 \tab 6270 \tab 21472 \cr
}
}

\seealso{
Other War and International: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{commun}},
\code{\link{commun10}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{japan}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}

Other Countries: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{japan}},
\code{\link{russia}}

Other Split Ballots: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{aidsads}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blkcol}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{colaff}},
\code{\link{colop}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{father}},
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fewrite}},
\code{\link{forbdcom}},
\code{\link{forbdmar}},
\code{\link{forbdrac}},
\code{\link{form}},
\code{\link{formwt}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{happy7}},
\code{\link{healer}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{japan}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{king}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{liberatr}},
\code{\link{lover}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mapa}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{meserve}},
\code{\link{mother}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{opfamily}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racimp}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{rank}()},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{russia}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{spouse}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{uswar}},
\code{\link{uswary}},
\code{\link{xmovie1}}
}
\concept{Countries}
\concept{Split Ballots}
\concept{War and International}
\keyword{variable}
