% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{issp}
\alias{issp}
\title{Filter for issp cases}
\source{
General Social Survey https://gss.norc.org
}
\description{
issp
}
\details{
Question Status on International Social Survey Program Self-Completion Supplement.
}
\section{Values}{

\itemize{
\item \code{1} r did issp
\item \code{2} r did not do issp
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2008 \tab B/C/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5576/vshow}{GSS Data Explorer page for \code{issp}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{formwt}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{version}}
}
\concept{Sampling and Weights}
\keyword{variable}
