% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{iwrkhard}
\alias{iwrkhard}
\title{How hard does r work at job}
\source{
General Social Survey https://gss.norc.org
}
\description{
iwrkhard
}
\details{
Question Which of the following statements best describes your feelings about the job?
}
\section{Values}{

\itemize{
\item \code{1} only minimum
\item \code{2} not to interfere
\item \code{3} real hard
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4531/vshow}{GSS Data Explorer page for \code{iwrkhard}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab only minimum \tab not to interfere \tab real hard \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 65 \tab 265 \tab 522 \tab 11 \tab 35 \tab 1537 \cr
   1998 \tab 1967 \tab 67 \tab 306 \tab 451 \tab 13 \tab 28 \tab 2832 \cr
   Total \tab 2606 \tab 132 \tab 571 \tab 973 \tab 24 \tab 63 \tab 4369 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
