% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{jew16}
\alias{jew16}
\title{What sort of jew r was at age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
jew16
}
\details{
Question Were you raised as a Jew?
A. Were you Orthodox, Conservative, Reform, or none of these? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \code{1} orthodox
\item \code{2} conservative
\item \code{3} reform
\item \code{4} none of these
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/345/vshow}{GSS Data Explorer page for \code{jew16}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab orthodox \tab conservative \tab reform \tab none of these \tab no answer \tab don't know \tab other \tab skipped on web \tab Total \cr
   1988 \tab 1450 \tab 7 \tab 14 \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1505 \tab 6 \tab 10 \tab 7 \tab 4 \tab 5 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1337 \tab 6 \tab 11 \tab 7 \tab 2 \tab 9 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1480 \tab 4 \tab 12 \tab 13 \tab 3 \tab 5 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1571 \tab 3 \tab 11 \tab 13 \tab 7 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2930 \tab 4 \tab 22 \tab 20 \tab 8 \tab 8 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2836 \tab 8 \tab 25 \tab 29 \tab 6 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2752 \tab 8 \tab 22 \tab 17 \tab 3 \tab 29 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 2739 \tab 2 \tab 19 \tab 20 \tab 11 \tab 26 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2694 \tab 4 \tab 17 \tab 26 \tab 8 \tab 15 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 2733 \tab 14 \tab 14 \tab 19 \tab 12 \tab 20 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4454 \tab 16 \tab 12 \tab 16 \tab 11 \tab - \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 1974 \tab 10 \tab 13 \tab 13 \tab 3 \tab 10 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2003 \tab 4 \tab 12 \tab 11 \tab 6 \tab 8 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1942 \tab 2 \tab 6 \tab 12 \tab 7 \tab 5 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2485 \tab 6 \tab 12 \tab 15 \tab 8 \tab 12 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2802 \tab 8 \tab 17 \tab 17 \tab 12 \tab 10 \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2299 \tab 4 \tab 10 \tab 19 \tab 8 \tab 8 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3941 \tab 6 \tab 22 \tab 38 \tab 18 \tab 3 \tab - \tab 3 \tab 1 \tab 4032 \cr
   2022 \tab 3431 \tab 6 \tab 14 \tab 23 \tab 9 \tab 18 \tab 20 \tab 3 \tab 20 \tab 3544 \cr
   Total \tab 49358 \tab 128 \tab 295 \tab 344 \tab 147 \tab 192 \tab 23 \tab 7 \tab 21 \tab 50515 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom16}},
\code{\link{fund16}},
\code{\link{jew16aj}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}
}
\concept{Core}
\keyword{variable}
