% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{jew_next}
\alias{jew_next}
\title{What sort of jew - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
jew_next
}
\details{
Question  Do you consider yourself Orthodox, Conservative, Reform or none of these?
}
\section{Values}{

\itemize{
\item \verb{[1]} orthodox
\item \verb{[2]} conservative
\item \verb{[3]} reform
\item \verb{[4]} none of these
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8003/vshow}{GSS Data Explorer page for \code{jew_next}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab orthodox \tab conservative \tab reform \tab none of these \tab other \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1190 \tab 5 \tab 7 \tab 15 \tab 9 \tab 5 \tab 2308 \tab 5 \tab 3544 \cr
   Total \tab 1190 \tab 5 \tab 7 \tab 15 \tab 9 \tab 5 \tab 2308 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
