% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{jewaj}
\alias{jewaj}
\title{What sort of jew r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
jewaj
}
\details{
Question B. Do you consider yourself Orthodox, Conservative, Reform or none of these? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \verb{[1]} orthodox
\item \verb{[2]} conservative
\item \verb{[3]} reform
\item \verb{[4]} none of these
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/292/vshow}{GSS Data Explorer page for \code{jewaj}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab conservative \tab reform \tab none of these \tab no answer \tab Total \cr
   2012 \tab 1954 \tab 2 \tab 2 \tab 8 \tab 8 \tab 1974 \cr
   Total \tab 1954 \tab 2 \tab 2 \tab 8 \tab 8 \tab 1974 \cr
}
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jew}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{spattend}}
}
\concept{Core}
\keyword{variable}
