% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{jews}
\alias{jews}
\title{What is your personal attitude towards members of the following religious group}
\source{
General Social Survey https://gss.norc.org
}
\description{
jews
}
\details{
Question What is your personal attitude towards members of the following religious groups? Jews
}
\section{Values}{

\itemize{
\item \verb{[1]} very positive
\item \verb{[2]} somewhat positive
\item \verb{[3]} neither positive nor negative
\item \verb{[4]} somewhat negative
\item \verb{[5]} very negative
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7290/vshow}{GSS Data Explorer page for \code{jews}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very positive \tab somewhat positive \tab neither positive nor negative \tab somewhat negative \tab very negative \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1173 \tab 245 \tab 361 \tab 462 \tab 43 \tab 15 \tab 32 \tab 17 \tab 2348 \cr
   Total \tab 1173 \tab 245 \tab 361 \tab 462 \tab 43 \tab 15 \tab 32 \tab 17 \tab 2348 \cr
}
}

\keyword{variable}
