% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_259.R
\name{jewsschl}
\alias{jewsschl}
\title{Does r know jews from school}
\source{
General Social Survey https://gss.norc.org
}
\description{
jewsschl
}
\details{
Question Now, I’m going to ask some questions about these people you personally know.
A. Do you know any of these Jews from when you went to school or college?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2345/vshow}{GSS Data Explorer page for \code{jewsschl}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2015 \tab 331 \tab 391 \tab 6 \tab 74 \tab 2817 \cr
   Total \tab 2015 \tab 331 \tab 391 \tab 6 \tab 74 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkwrk}},
\code{\link{hispschl}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewswrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
