% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{jewtemp}
\alias{jewtemp}
\title{Feeling thermometer:jews}
\source{
General Social Survey https://gss.norc.org
}
\description{
jewtemp
}
\details{
Question I'd like to get your feelings toward groups that are in the news these days. I will use something we call the feeling thermometer, and here is how it works: I'll read the names of a group and I'd like you to rate that group using the feeling thermometer. Ratings between 50 degrees and 100 degrees mean that you feel favorable and warm toward the group. Ratings between 0 degrees and 50 degrees mean that you don't feel favorable toward the group and that you don't care too much for that group. If we come to a group whose name you Don't recognize, you don't need to rate that group. Just tell me and we'll move on to the next one. IF you do recognize the name, but you don't feel particularly warm or cold toward the group, you would rate the group at the 50 degree mark. How would you rate this group using the thermometer?
E. Jews
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   B/C/- \tab 2004 \cr
   Full \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/377/vshow}{GSS Data Explorer page for \code{jewtemp}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 15 \tab 20 \tab 25 \tab 30 \tab 35 \tab 40 \tab 45 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 55 \tab 60 \tab 70 \tab 75 \tab 78 \tab 80 \tab 85 \tab 90 \tab 95 \tab don't know \tab no answer \tab 65 \tab 83 \tab 98 \tab 99 \tab 17 \tab 33 \tab 6 \tab 72 \tab 2 \tab 7 \tab 77 \tab 82 \tab 86 \tab Total \cr
   1986 \tab - \tab 33 \tab 1 \tab 12 \tab 110 \tab 12 \tab 6 \tab 4 \tab 38 \tab 1 \tab 88 \tab 5 \tab 2 \tab 1 \tab 260 \tab 6 \tab 1 \tab 6 \tab 261 \tab 300 \tab 18 \tab 1 \tab 25 \tab 185 \tab 9 \tab 1 \tab 65 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1988 \tab - \tab 27 \tab - \tab 6 \tab 93 \tab 20 \tab 9 \tab 6 \tab 35 \tab 1 \tab 56 \tab 2 \tab - \tab - \tab 511 \tab - \tab - \tab - \tab 163 \tab 220 \tab 17 \tab 1 \tab 22 \tab 174 \tab 8 \tab 1 \tab 83 \tab 20 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 531 \tab 16 \tab 1 \tab 4 \tab 80 \tab 14 \tab 8 \tab 1 \tab 26 \tab 2 \tab 55 \tab - \tab - \tab - \tab 274 \tab - \tab - \tab 2 \tab 120 \tab 162 \tab 17 \tab - \tab 8 \tab 113 \tab 5 \tab 2 \tab 76 \tab 14 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   2004 \tab 1939 \tab 7 \tab 3 \tab 3 \tab 58 \tab 9 \tab 1 \tab 4 \tab 8 \tab 1 \tab 18 \tab 1 \tab - \tab 5 \tab 352 \tab - \tab - \tab 2 \tab 78 \tab 109 \tab 19 \tab - \tab 19 \tab 85 \tab 9 \tab 2 \tab 47 \tab 19 \tab 7 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2812 \cr
   Total \tab 2470 \tab 83 \tab 5 \tab 25 \tab 341 \tab 55 \tab 24 \tab 15 \tab 107 \tab 5 \tab 217 \tab 8 \tab 2 \tab 6 \tab 1397 \tab 6 \tab 1 \tab 10 \tab 622 \tab 791 \tab 71 \tab 2 \tab 74 \tab 557 \tab 31 \tab 6 \tab 271 \tab 72 \tab 12 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 7300 \cr
}
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{fepriest}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}},
\code{\link{relgrade}}
}
\concept{Core}
\keyword{variable}
