% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{jobeasy}
\alias{jobeasy}
\title{How easy if r find a job as good as current one}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobeasy
}
\details{
Question A. How difficult or easy do you think it would be for you to find a job at least as good as your current one?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} fairly easy
\item \code{3} neither easy nor difficult
\item \code{4} fairly difficult
\item \code{5} very difficult
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4618/vshow}{GSS Data Explorer page for \code{jobeasy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very easy \tab fairly easy \tab neither easy nor difficult \tab fairly difficult \tab very difficult \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 158 \tab 292 \tab 175 \tab 266 \tab 120 \tab 4 \tab 3 \tab 4510 \cr
   2016 \tab 1927 \tab 120 \tab 282 \tab 158 \tab 266 \tab 107 \tab 2 \tab 5 \tab 2867 \cr
   Total \tab 5419 \tab 278 \tab 574 \tab 333 \tab 532 \tab 227 \tab 6 \tab 8 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
