% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{jobfind1}
\alias{jobfind1}
\title{How easy for r to find a same job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobfind1
}
\details{
Question How easy would it be for you to find a job with another employer with approximately the same income and fringe benefits as you have now?
}
\section{Values}{

\itemize{
\item \code{1} very easy to find similar job
\item \code{2} somewhat easy to find similar job
\item \code{3} not easy at all to find similar job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2819/vshow}{GSS Data Explorer page for \code{jobfind1}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}},
\code{\link{wksexism}}
}
\concept{Quality of Working Life}
\keyword{variable}
