% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{jobfrnds}
\alias{jobfrnds}
\title{Times r sought job information from friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobfrnds
}
\details{
Question Now I'd like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit: Click on a link in an email message from a friend, relative, or co-worker.
F. Friends outside of work or relatives?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2615/vshow}{GSS Data Explorer page for \code{jobfrnds}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab 6 or more times \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2220 \tab 199 \tab 199 \tab 88 \tab 52 \tab 1 \tab 6 \tab 2765 \cr
   2004 \tab 2453 \tab 135 \tab 117 \tab 55 \tab 44 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 4673 \tab 334 \tab 316 \tab 143 \tab 96 \tab 2 \tab 13 \tab 5577 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emailyr}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webable}},
\code{\link{webyr}}
}
\concept{Information Society}
\keyword{variable}
