% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{jobhelp}
\alias{jobhelp}
\title{Can help others}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobhelp
}
\details{
Question On these cards there are various aspects of jobs. How important Do you personally consider these job characteristics? Using the scale from one to seven for your answers again, tell me for each card its letter and the number you've decided on.
SHUFFLE THE PINK CARDS BEFORE HANDING TO RESPONDENT.
J. An occupation in which one can help others
}
\section{Values}{

\itemize{
\item \code{1} unimportant
\item \code{7} very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/564/vshow}{GSS Data Explorer page for \code{jobhelp}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab unimportant \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab don't know \tab no answer \tab Total \cr
   1982 \tab 923 \tab 14 \tab 16 \tab 29 \tab 50 \tab 103 \tab 200 \tab 506 \tab 6 \tab 13 \tab 1860 \cr
   Total \tab 923 \tab 14 \tab 16 \tab 29 \tab 50 \tab 103 \tab 200 \tab 506 \tab 6 \tab 13 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{jobaccmp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()},
\code{\link{rank10}}
}
\concept{Core}
\keyword{variable}
