% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{jobhrw}
\alias{jobhrw}
\title{Hours per week using work computer job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobhrw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
A. Your paid job?
2. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2568/vshow}{GSS Data Explorer page for \code{jobhrw}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 15 \tab 16 \tab 2 \tab 20 \tab 22 \tab 23 \tab 24 \tab 25 \tab 28 \tab 3 \tab 30 \tab 32 \tab 34 \tab 35 \tab 36 \tab 38 \tab 4 \tab 40 \tab 45 \tab 5 \tab 50 \tab 52 \tab 55 \tab 6 \tab 60 \tab 7 \tab 8 \tab no answer \tab Total \cr
   2000 \tab 2486 \tab 19 \tab 18 \tab 22 \tab 3 \tab 7 \tab 1 \tab 19 \tab 26 \tab 2 \tab 2 \tab 2 \tab 10 \tab 2 \tab 7 \tab 16 \tab 1 \tab 2 \tab 9 \tab 1 \tab 3 \tab 10 \tab 44 \tab 3 \tab 14 \tab 3 \tab 1 \tab 1 \tab 8 \tab 7 \tab 2 \tab 10 \tab 56 \tab 2817 \cr
   Total \tab 2486 \tab 19 \tab 18 \tab 22 \tab 3 \tab 7 \tab 1 \tab 19 \tab 26 \tab 2 \tab 2 \tab 2 \tab 10 \tab 2 \tab 7 \tab 16 \tab 1 \tab 2 \tab 9 \tab 1 \tab 3 \tab 10 \tab 44 \tab 3 \tab 14 \tab 3 \tab 1 \tab 1 \tab 8 \tab 7 \tab 2 \tab 10 \tab 56 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compwork}},
\code{\link{emhrh}},
\code{\link{emminh}},
\code{\link{jobminw}},
\code{\link{perhrh}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwhrh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
