% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{jobinc}
\alias{jobinc}
\title{High income}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobinc
}
\details{
Question A. Would you please look at this card and tell me which one thing on this list you would most prefer in a job?
B. Which comes next?
C. Which is third most important?
D. Which is fourth most important?
A. High Income
}
\section{Values}{

\itemize{
\item \verb{[1]} most impt
\item \verb{[2]} second
\item \verb{[3]} third
\item \verb{[4]} fourth
\item \verb{[5]} fifth
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/B \tab 2012, 2014 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/550/vshow}{GSS Data Explorer page for \code{jobinc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most impt \tab second \tab third \tab fourth \tab fifth \tab no answer \tab Total \cr
   1973 \tab - \tab 272 \tab 352 \tab 450 \tab 286 \tab 102 \tab 42 \tab 1504 \cr
   1974 \tab - \tab 269 \tab 301 \tab 451 \tab 303 \tab 130 \tab 30 \tab 1484 \cr
   1976 \tab - \tab 295 \tab 326 \tab 469 \tab 281 \tab 78 \tab 50 \tab 1499 \cr
   1977 \tab - \tab 304 \tab 362 \tab 458 \tab 260 \tab 96 \tab 50 \tab 1530 \cr
   1980 \tab - \tab 288 \tab 383 \tab 463 \tab 235 \tab 67 \tab 32 \tab 1468 \cr
   1982 \tab 937 \tab 260 \tab 226 \tab 214 \tab 135 \tab 47 \tab 41 \tab 1860 \cr
   1984 \tab - \tab 276 \tab 357 \tab 475 \tab 272 \tab 61 \tab 32 \tab 1473 \cr
   1985 \tab - \tab 291 \tab 386 \tab 466 \tab 293 \tab 64 \tab 34 \tab 1534 \cr
   1987 \tab - \tab 436 \tab 466 \tab 525 \tab 254 \tab 91 \tab 47 \tab 1819 \cr
   1988 \tab 504 \tab 200 \tab 228 \tab 314 \tab 160 \tab 44 \tab 31 \tab 1481 \cr
   1989 \tab 504 \tab 210 \tab 245 \tab 314 \tab 174 \tab 37 \tab 53 \tab 1537 \cr
   1990 \tab 455 \tab 182 \tab 231 \tab 266 \tab 152 \tab 49 \tab 37 \tab 1372 \cr
   1991 \tab 524 \tab 228 \tab 233 \tab 264 \tab 169 \tab 51 \tab 48 \tab 1517 \cr
   1993 \tab 531 \tab 202 \tab 297 \tab 289 \tab 189 \tab 60 \tab 38 \tab 1606 \cr
   1994 \tab 2508 \tab 104 \tab 121 \tab 138 \tab 81 \tab 19 \tab 21 \tab 2992 \cr
   2006 \tab 2992 \tab 356 \tab 379 \tab 415 \tab 248 \tab 100 \tab 20 \tab 4510 \cr
   2012 \tab 1002 \tab 257 \tab 261 \tab 221 \tab 144 \tab 64 \tab 25 \tab 1974 \cr
   2014 \tab 1288 \tab 337 \tab 344 \tab 315 \tab 167 \tab 71 \tab 16 \tab 2538 \cr
   Total \tab 11245 \tab 4767 \tab 5498 \tab 6507 \tab 3803 \tab 1231 \tab 647 \tab 33698 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{jobfind}},
\code{\link{joblose}},
\code{\link{jobsec}},
\code{\link{richwork}},
\code{\link{satjob}}
}
\concept{Core}
\keyword{variable}
