% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{jobinfo}
\alias{jobinfo}
\title{R sought employment info past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobinfo
}
\details{
Question Now I'd like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit: Click on a link in an email message from a friend, relative, or co-worker.
A. In the past year, that is, between (INSERT CURRENT MONTH) 2001/2003 and  (INSERT CURRENT MONTH) 2002/2004"have you searched for information about a new job or explored career opportunities?
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2610/vshow}{GSS Data Explorer page for \code{jobinfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab 1 or 2 times \tab 3-5 times \tab 6 or more times \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 842 \tab 223 \tab 146 \tab 170 \tab 6 \tab 6 \tab 2765 \cr
   2004 \tab 1828 \tab 624 \tab 78 \tab 92 \tab 184 \tab 1 \tab 5 \tab 2812 \cr
   Total \tab 3200 \tab 1466 \tab 301 \tab 238 \tab 354 \tab 7 \tab 11 \tab 5577 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{byemprob}},
\code{\link{intaccss}},
\code{\link{jobpaper}},
\code{\link{numprobs}}
}
\concept{Information Society}
\keyword{variable}
