% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{jobminw}
\alias{jobminw}
\title{Minutes per week using work computer job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobminw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
A. Your paid job?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2567/vshow}{GSS Data Explorer page for \code{jobminw}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 15 \tab 20 \tab 3 \tab 30 \tab 40 \tab 45 \tab 50 \tab 8 \tab no answer \tab Total \cr
   2000 \tab 2486 \tab 261 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 1 \tab 46 \tab 2817 \cr
   Total \tab 2486 \tab 261 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 1 \tab 46 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compwork}},
\code{\link{emhrh}},
\code{\link{emminh}},
\code{\link{jobhrw}},
\code{\link{perhrh}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwhrh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
