% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{jobplace}
\alias{jobplace}
\title{Times r sought job information from counseling service}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobplace
}
\details{
Question Now I'd like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit: Click on a link in an email message from a friend, relative, or co-worker.
G. Any job placement of career counseling service?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2616/vshow}{GSS Data Explorer page for \code{jobplace}}.
}

\seealso{
Other Information Society: 
\code{\link{emailyr}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webable}},
\code{\link{webyr}}
}
\concept{Information Society}
\keyword{variable}
