% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_62.R
\name{jobpromo}
\alias{jobpromo}
\title{Chances for advancement}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobpromo
}
\details{
Question A. Would you please look at this card and tell me which one thing on this list you would most prefer in a job?
B. Which comes next?
C. Which is third most important?
D. Which is fourth most important?
D. Chances for advancement
}
\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/553/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab fifth \tab fourth \tab most impt \tab no answer \tab second \tab third \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 142 \tab 195 \tab 266 \tab 42 \tab 517 \tab 342 \tab - \tab 1504 \cr
   1974 \tab - \tab 119 \tab 193 \tab 261 \tab 30 \tab 549 \tab 332 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 138 \tab 217 \tab 257 \tab 50 \tab 529 \tab 308 \tab - \tab 1499 \cr
   1977 \tab - \tab 131 \tab 204 \tab 296 \tab 50 \tab 507 \tab 342 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 105 \tab 194 \tab 272 \tab 32 \tab 502 \tab 363 \tab - \tab 1468 \cr
   1982 \tab 937 \tab 95 \tab 182 \tab 156 \tab 41 \tab 256 \tab 193 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 108 \tab 197 \tab 277 \tab 32 \tab 516 \tab 343 \tab - \tab 1473 \cr
   1985 \tab - \tab 87 \tab 189 \tab 334 \tab 34 \tab 546 \tab 344 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 134 \tab 265 \tab 334 \tab 47 \tab 615 \tab 424 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 80 \tab 132 \tab 191 \tab 31 \tab 322 \tab 221 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 66 \tab 139 \tab 170 \tab 53 \tab 361 \tab 244 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 66 \tab 127 \tab 156 \tab 37 \tab 312 \tab 219 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 79 \tab 152 \tab 154 \tab 48 \tab 329 \tab 231 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 82 \tab 163 \tab 177 \tab 38 \tab 358 \tab 257 \tab - \tab 1606 \cr
   1994 \tab 2508 \tab 41 \tab 77 \tab 82 \tab 21 \tab 149 \tab 114 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2992 \tab 159 \tab 265 \tab 283 \tab 20 \tab 439 \tab 352 \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1002 \tab 158 \tab 194 \tab 115 \tab 26 \tab 246 \tab 233 \tab - \tab 1974 \cr
   2014 \tab 1288 \tab 196 \tab 248 \tab 163 \tab 18 \tab 325 \tab 300 \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 46393 \tab 1986 \tab 3333 \tab 3944 \tab 650 \tab 7378 \tab 5162 \tab 6853 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   2006 \tab D/- \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
}
}

\seealso{
Other Core: 
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobrise}}

Other Job: 
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobrise}}

Other Split Ballots: 
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobrise}}

Other Work: 
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobrise}}
}
\concept{Core}
\concept{Job}
\concept{Split Ballots}
\concept{Work}
\keyword{variable}
