% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{jobresp}
\alias{jobresp}
\title{Responsible job tasks}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobresp
}
\details{
Question On these cards there are various aspects of jobs. How important Do you personally consider these job characteristics? Using the scale from one to seven for your answers again, tell me for each card its letter and the number you've decided on.
SHUFFLE THE PINK CARDS BEFORE HANDING TO RESPONDENT.
H. Responsible job tasks
}
\section{Values}{

\itemize{
\item \code{1} unimportant
\item \code{7} very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/562/vshow}{GSS Data Explorer page for \code{jobresp}}.
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()},
\code{\link{rank10}}

Other Split Ballots: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()}
}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
