% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_62.R
\name{jobrise}
\alias{jobrise}
\title{Opportunity for advancement}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobrise
}
\details{
Question On these cards there are various aspects of jobs. How important Do you personally consider these job characteristics? Using the scale from one to seven for your answers again, tell me for each card its letter and the number you've decided on.
SHUFFLE THE PINK CARDS BEFORE HANDING TO RESPONDENT.
C. Good opportunities for advancement
}
\section{Values}{

\itemize{
\item \verb{[1]} unimportant
\item \verb{[7]} very important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/557/vshow}{GSS Data Explorer page for \code{jobrise}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab unimportant \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab don't know \tab no answer \tab Total \cr
   1982 \tab 923 \tab 15 \tab 10 \tab 21 \tab 44 \tab 124 \tab 221 \tab 480 \tab 6 \tab 16 \tab 1860 \cr
   Total \tab 923 \tab 15 \tab 10 \tab 21 \tab 44 \tab 124 \tab 221 \tab 480 \tab 6 \tab 16 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpromo}}
}
\concept{Core}
\keyword{variable}
