% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{jobsat}
\alias{jobsat}
\title{How satisfied is r with his/her job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobsat
}
\details{
Question How satisfied are you in your job?
}
\section{Values}{

\itemize{
\item \code{1} completely satisfied
\item \code{2} very satisfied
\item \code{3} fairly satisfied
\item \code{4} neither satisfied nor dissatisfied
\item \code{5} fairly dissatisfied
\item \code{6} very dissatisfied
\item \code{7} completely dissatisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4561/vshow}{GSS Data Explorer page for \code{jobsat}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab completely satisfied \tab very satisfied \tab fairly satisfied \tab neither satisfied nor dissatisfied \tab fairly dissatisfied \tab very dissatisfied \tab completely dissatisfied \tab don't know \tab no answer \tab Total \cr
   1989 \tab 639 \tab 141 \tab 296 \tab 323 \tab 52 \tab 45 \tab 11 \tab 5 \tab 5 \tab 20 \tab 1537 \cr
   1998 \tab 1967 \tab 133 \tab 271 \tab 292 \tab 61 \tab 53 \tab 15 \tab 9 \tab 3 \tab 28 \tab 2832 \cr
   2006 \tab 3492 \tab 187 \tab 381 \tab 304 \tab 69 \tab 31 \tab 31 \tab 13 \tab - \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 164 \tab 343 \tab 293 \tab 54 \tab 52 \tab 20 \tab 7 \tab 1 \tab 6 \tab 2867 \cr
   Total \tab 8025 \tab 625 \tab 1291 \tab 1212 \tab 236 \tab 181 \tab 77 \tab 34 \tab 9 \tab 56 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
