% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{jobskill}
\alias{jobskill}
\title{How important training in developing skills}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobskill
}
\details{
Question Now think about the skills that you actually use in your job. How important would you say each of the following was in developing these skills?
B. Training or experience gained in my present or a previous job
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} not impt/unimp
\item \code{4} not important
\item \code{5} not imprt at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4584/vshow}{GSS Data Explorer page for \code{jobskill}}.
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}

Other Job: 
\code{\link{absent}},
\code{\link{edcskill}},
\code{\link{useskill}},
\code{\link{wktenure}}

Other Work: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Job}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
