% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{jobsoc}
\alias{jobsoc}
\title{Useful to society}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobsoc
}
\details{
Question On these cards there are various aspects of jobs. How important Do you personally consider these job characteristics? Using the scale from one to seven for your answers again, tell me for each card its letter and the number you've decided on.
SHUFFLE THE PINK CARDS BEFORE HANDING TO RESPONDENT.
K. A job that is useful to society
}
\section{Values}{

\itemize{
\item \verb{[1]} unimportant
\item \verb{[7]} very important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/565/vshow}{GSS Data Explorer page for \code{jobsoc}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab unimportant \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab don't know \tab no answer \tab Total \cr
   1982 \tab 923 \tab 23 \tab 17 \tab 30 \tab 62 \tab 130 \tab 209 \tab 443 \tab 7 \tab 16 \tab 1860 \cr
   Total \tab 923 \tab 23 \tab 17 \tab 30 \tab 62 \tab 130 \tab 209 \tab 443 \tab 7 \tab 16 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{rank}()},
\code{\link{rank10}}
}
\concept{Core}
\keyword{variable}
