% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{jobtrain}
\alias{jobtrain}
\title{Obligation to train for civilian jobs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobtrain
}
\details{
Question Most people in the Armed Forces are taught skills they can use in civilian jobs later. But some don't get such training. They are taught only combat skills. Do you think the Armed Forces have an obligation to train everybody in service For civilian jobs later, or is that not a responsibility of the Armed Forces?
}
\section{Values}{

\itemize{
\item \code{1} yes, obligation
\item \code{2} not obligation
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/821/vshow}{GSS Data Explorer page for \code{jobtrain}}.
}

\seealso{
Other Core: 
\code{\link{defwrkev}},
\code{\link{guerilla}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()},
\code{\link{vetaid}}
}
\concept{Core}
\keyword{variable}
