% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{jobvsfa1}
\alias{jobvsfa1}
\title{How often jobtakes too much time to fulfill family resp}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobvsfa1
}
\details{
Question How often has each of the following happened to you during the past three months?
B. It has been difficult for you to fulfil your family responsibilities because of the amount of time you spent on your job.
}
\section{Values}{

\itemize{
\item \verb{[0]} doesn't apply/no job
\item \verb{[1]} several times a week
\item \verb{[2]} several times a month
\item \verb{[3]} once or twice
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4476/vshow}{GSS Data Explorer page for \code{jobvsfa1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab doesn't apply/no job \tab several times a week \tab several times a month \tab once or twice \tab never \tab don't know \tab no answer \tab 5 \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 474 \tab 105 \tab 132 \tab 213 \tab 366 \tab 4 \tab 8 \tab - \tab - \tab 1974 \cr
   2022 \tab 2386 \tab - \tab 116 \tab 161 \tab 234 \tab 297 \tab 7 \tab 5 \tab 330 \tab 8 \tab 3544 \cr
   Total \tab 3058 \tab 474 \tab 221 \tab 293 \tab 447 \tab 663 \tab 11 \tab 13 \tab 330 \tab 8 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
