% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{jobvshme}
\alias{jobvshme}
\title{How often your family interfered with your job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobvshme
}
\details{
Question In the last few years, how often would you say that your family or household responsibilities have interfered with  your ability to devote full attention to your job? Does this happen frequently, sometimes, seldom or never?
}
\section{Values}{

\itemize{
\item \verb{[1]} frequently
\item \verb{[2]} sometimes
\item \verb{[3]} seldom
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1364/vshow}{GSS Data Explorer page for \code{jobvshme}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab frequently \tab sometimes \tab seldom \tab never \tab no answer \tab Total \cr
   1991 \tab 605 \tab 42 \tab 158 \tab 293 \tab 407 \tab 12 \tab 1517 \cr
   Total \tab 605 \tab 42 \tab 158 \tab 293 \tab 407 \tab 12 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{schoolng}},
\code{\link{unmanrel}},
\code{\link{unvote}}
}
\concept{Work Organizations}
\keyword{variable}
