% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{jobyears}
\alias{jobyears}
\title{For how long have you had your present job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobyears
}
\details{
Question For about how long have you had your present job?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1291/vshow}{GSS Data Explorer page for \code{jobyears}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 1.16 \tab 1.25 \tab 1.33 \tab 1.41 \tab 1.5 \tab 1.83 \tab 10 \tab 11 \tab 12 \tab 12.08 \tab 13 \tab 14 \tab 15 \tab 15.25 \tab 16 \tab 16.41 \tab 17 \tab 17.5 \tab 18 \tab 19 \tab 2 \tab 2.5 \tab 2.83 \tab 20 \tab 22 \tab 23 \tab 24 \tab 25 \tab 3 \tab 3.33 \tab 3.5 \tab 3.66 \tab 3.75 \tab 30 \tab 33 \tab 4 \tab 4.75 \tab 41 \tab 48.25 \tab 5 \tab 50 \tab 58 \tab 6 \tab 6.41 \tab 6.5 \tab 66 \tab 7 \tab 7.66 \tab 7.75 \tab 75 \tab 8 \tab 8.5 \tab 83 \tab 9 \tab 91 \tab Total \cr
   1991 \tab 1252 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 7 \tab 3 \tab 6 \tab 1 \tab 2 \tab 6 \tab 5 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 22 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 6 \tab 17 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 9 \tab 16 \tab 1 \tab 4 \tab 1 \tab 19 \tab 10 \tab 3 \tab 8 \tab 1 \tab 1 \tab 7 \tab 12 \tab 1 \tab 1 \tab 3 \tab 17 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1517 \cr
   Total \tab 1252 \tab 17 \tab 2 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 7 \tab 3 \tab 6 \tab 1 \tab 2 \tab 6 \tab 5 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 22 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 6 \tab 17 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 9 \tab 16 \tab 1 \tab 4 \tab 1 \tab 19 \tab 10 \tab 3 \tab 8 \tab 1 \tab 1 \tab 7 \tab 12 \tab 1 \tab 1 \tab 3 \tab 17 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{promotng}},
\code{\link{samejob}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}
}
\concept{Work Organizations}
\keyword{variable}
