% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_116.R
\name{join16}
\alias{join16}
\title{Was r ever confirmed as member of church}
\source{
General Social Survey https://gss.norc.org
}
\description{
join16
}
\details{
Question Did you ever join a church (synagogue) when you were growing up, that is, become a member by confirmation or such?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1071/vshow}{GSS Data Explorer page for \code{join16}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 1000 \tab 471 \tab 2 \tab 8 \tab 1481 \cr
   Total \tab 0 \tab 1000 \tab 471 \tab 2 \tab 8 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{god}},
\code{\link{grace16}},
\code{\link{lapsed}},
\code{\link{reborn}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{tvrelig}}
}
\concept{Religion}
\keyword{variable}
