% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_389.R
\name{journlsm}
\alias{journlsm}
\title{How scientific is journalism}
\source{
General Social Survey https://gss.norc.org
}
\description{
journlsm
}
\details{
Question B. Journalism. (Is journalism...?)
}
\section{Values}{

\itemize{
\item \code{1} very scientific
\item \code{2} pretty scientific
\item \code{3} not too scientific
\item \code{4} not scientific at all
\item \code{5} haven't heard of it (volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3538/vshow}{GSS Data Explorer page for \code{journlsm}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab haven't heard of it (volunteered) \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1469 \tab 30 \tab 73 \tab 245 \tab 127 \tab 3 \tab 25 \tab 2 \tab 1974 \cr
   Total \tab 1469 \tab 30 \tab 73 \tab 245 \tab 127 \tab 3 \tab 25 \tab 2 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{engint}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{farming}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
