% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_38.R
\name{judge}
\alias{judge}
\title{Image of god as judge}
\source{
General Social Survey https://gss.norc.org
}
\description{
judge
}
\details{
Question When you think about God, how likely are each of these images to come to your mind?  Would you say extremely likely, somewhat likely, not too likely, or not likely at all? (HAND CARD AD. READ EACH ITEM.) (CIRCLE ONE NUMBER FOR EACH WORD)
A. Judge
}
\section{Values}{

\itemize{
\item \code{1} extremely likely
\item \code{2} somewhat likely
\item \code{3} not too likely
\item \code{4} not likely at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/322/vshow}{GSS Data Explorer page for \code{judge}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab extremely likely \tab somewhat likely \tab not too likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab - \tab 731 \tab 470 \tab 208 \tab 148 \tab 25 \tab 17 \tab 1599 \cr
   1984 \tab 989 \tab 223 \tab 138 \tab 55 \tab 52 \tab 5 \tab 11 \tab 1473 \cr
   Total \tab 989 \tab 954 \tab 608 \tab 263 \tab 200 \tab 30 \tab 28 \tab 3072 \cr
}
}

\seealso{
Other Core: 
\code{\link{creator}},
\code{\link{father}},
\code{\link{friend}},
\code{\link{lover}},
\code{\link{master}},
\code{\link{mother}},
\code{\link{neargod}},
\code{\link{redeemer}},
\code{\link{spouse}}
}
\concept{Core}
\keyword{variable}
