% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{judgeart}
\alias{judgeart}
\title{Only a few people can judge art}
\source{
General Social Survey https://gss.norc.org
}
\description{
judgeart
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
A. Only a few people have the knowledge and ability to judge excellence in the arts.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1422/vshow}{GSS Data Explorer page for \code{judgeart}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 87 \tab 642 \tab 614 \tab 139 \tab 107 \tab 17 \tab 1606 \cr
   1998 \tab 1387 \tab 98 \tab 439 \tab 526 \tab 213 \tab 156 \tab 13 \tab 2832 \cr
   Total \tab 1387 \tab 185 \tab 1081 \tab 1140 \tab 352 \tab 263 \tab 30 \tab 4438 \cr
}
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{modpaint}},
\code{\link{pclit}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}}
}
\concept{Culture}
\keyword{variable}
