% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_54.R
\name{judgment}
\alias{judgment}
\title{Good sense and sound judgment}
\source{
General Social Survey https://gss.norc.org
}
\description{
judgment
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
5. That he has good sense and sound judgment
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 most desirable
\item \verb{[2]} 3 most desirable
\item \verb{[3]} not mentioned
\item \verb{[4]} 3 least desirable
\item \verb{[5]} 1 least desirable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1975, 1976, 1978, 1980, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/478/vshow}{GSS Data Explorer page for \code{judgment}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 268 \tab 300 \tab 818 \tab 99 \tab 15 \tab 4 \tab - \tab 1504 \cr
   1975 \tab - \tab 227 \tab 279 \tab 874 \tab 88 \tab 13 \tab 9 \tab - \tab 1490 \cr
   1976 \tab - \tab 288 \tab 305 \tab 777 \tab 88 \tab 32 \tab 9 \tab - \tab 1499 \cr
   1978 \tab - \tab 244 \tab 322 \tab 833 \tab 100 \tab 21 \tab 10 \tab 2 \tab 1532 \cr
   1980 \tab 975 \tab 80 \tab 123 \tab 255 \tab 24 \tab 7 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 281 \tab 301 \tab 865 \tab 113 \tab 19 \tab 20 \tab - \tab 1599 \cr
   Total \tab 975 \tab 1388 \tab 1630 \tab 4422 \tab 512 \tab 107 \tab 55 \tab 3 \tab 9092 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicable}},
\code{\link{clean}},
\code{\link{consider}},
\code{\link{control}},
\code{\link{honest}},
\code{\link{obeys}},
\code{\link{responsi}},
\code{\link{role}},
\code{\link{success}}
}
\concept{Core}
\keyword{variable}
