% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_420.R
\name{kd1jwoth}
\alias{kd1jwoth}
\title{R's child 1 considered jewish}
\source{
General Social Survey https://gss.norc.org
}
\description{
kd1jwoth
}
\details{
Question Is (CHILD’S NAME) considered Jewish for any reason?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} half/part
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3838/vshow}{GSS Data Explorer page for \code{kd1jwoth}}.
}

\seealso{
Other Jewish Identity: 
\code{\link{bmitzvah}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kidnum}},
\code{\link{majwoth}},
\code{\link{pajwoth}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{synmem}}
}
\concept{Jewish Identity}
\keyword{variable}
