% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_419.R
\name{kd1relig}
\alias{kd1relig}
\title{Religion of r's child 1}
\source{
General Social Survey https://gss.norc.org
}
\description{
kd1relig
}
\details{
Question In what religion is (CHILD'S NAME) being raised? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[6]} buddhism
\item \verb{[7]} hinduism
\item \verb{[8]} other eastern religions
\item \verb{[9]} muslim/islam
\item \verb{[10]} orthodox-christian
\item \verb{[11]} christian
\item \verb{[12]} native american
\item \verb{[13]} inter-nondenominational
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3830/vshow}{GSS Data Explorer page for \code{kd1relig}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab buddhism \tab hinduism \tab other eastern religions \tab muslim/islam \tab orthodox-christian \tab christian \tab inter-nondenominational \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1261 \tab 352 \tab 173 \tab 14 \tab 153 \tab 5 \tab 2 \tab 5 \tab 4 \tab 9 \tab 1 \tab 35 \tab 6 \tab 1 \tab 2 \tab 2023 \cr
   2012 \tab 1955 \tab 2 \tab - \tab 9 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   Total \tab 3216 \tab 354 \tab 173 \tab 23 \tab 160 \tab 6 \tab 2 \tab 5 \tab 4 \tab 9 \tab 1 \tab 35 \tab 6 \tab 1 \tab 2 \tab 3997 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{bmitzvah}},
\code{\link{kd2relig}},
\code{\link{kd3relig}},
\code{\link{kd4relig}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{synmem}}
}
\concept{Jewish Identity}
\keyword{variable}
