% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_419.R
\name{kd4relig}
\alias{kd4relig}
\title{Religion of r's child 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
kd4relig
}
\details{
Question In what religion is (CHILD'S NAME) being raised? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[6]} buddhism
\item \verb{[7]} hinduism
\item \verb{[8]} other eastern religions
\item \verb{[9]} muslim/islam
\item \verb{[10]} orthodox-christian
\item \verb{[11]} christian
\item \verb{[12]} native american
\item \verb{[13]} inter-nondenominational
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3833/vshow}{GSS Data Explorer page for \code{kd4relig}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab buddhism \tab muslim/islam \tab christian \tab no answer \tab Total \cr
   2008 \tab 1960 \tab 33 \tab 13 \tab 3 \tab 9 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2023 \cr
   Total \tab 1960 \tab 33 \tab 13 \tab 3 \tab 9 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{bmitzvah}},
\code{\link{kd1relig}},
\code{\link{kd2relig}},
\code{\link{kd3relig}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{synmem}}
}
\concept{Jewish Identity}
\keyword{variable}
