% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_162.R
\name{kdalive2}
\alias{kdalive2}
\title{R 2nd child alive?}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdalive2
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
E2. Is second child alive?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1520/vshow}{GSS Data Explorer page for \code{kdalive2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2105 \tab 847 \tab 23 \tab 1 \tab 16 \tab 2992 \cr
   Total \tab 2105 \tab 847 \tab 23 \tab 1 \tab 16 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}}
}
\concept{Family Mobility}
\keyword{variable}
