% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_162.R
\name{kdalive5}
\alias{kdalive5}
\title{R 5th child alive?}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdalive5
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
E5. Is fifth child alive?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1523/vshow}{GSS Data Explorer page for \code{kdalive5}}.
}

\seealso{
Other Family Mobility: 
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}}
}
\concept{Family Mobility}
\keyword{variable}
