% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_162.R
\name{kdeduc1}
\alias{kdeduc1}
\title{Highest schl grade of r 1st child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdeduc1
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
F1. Highest school grade first child completed.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1528/vshow}{GSS Data Explorer page for \code{kdeduc1}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 4 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2332 \tab 4 \tab 1 \tab 13 \tab 16 \tab 258 \tab 29 \tab 74 \tab 28 \tab 120 \tab 13 \tab 17 \tab 1 \tab 13 \tab 1 \tab 1 \tab 1 \tab 5 \tab 7 \tab 8 \tab 50 \tab 2992 \cr
   Total \tab 2332 \tab 4 \tab 1 \tab 13 \tab 16 \tab 258 \tab 29 \tab 74 \tab 28 \tab 120 \tab 13 \tab 17 \tab 1 \tab 13 \tab 1 \tab 1 \tab 1 \tab 5 \tab 7 \tab 8 \tab 50 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc2}}
}
\concept{Family Mobility}
\keyword{variable}
