% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_162.R
\name{kdeduc2}
\alias{kdeduc2}
\title{Highest schl grade of r 2nd child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdeduc2
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
F2. Highest school grade second child completed.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1529/vshow}{GSS Data Explorer page for \code{kdeduc2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2474 \tab 7 \tab 1 \tab 22 \tab 6 \tab 204 \tab 38 \tab 47 \tab 22 \tab 86 \tab 9 \tab 17 \tab 4 \tab 6 \tab 9 \tab 2 \tab 6 \tab 32 \tab 2992 \cr
   Total \tab 2474 \tab 7 \tab 1 \tab 22 \tab 6 \tab 204 \tab 38 \tab 47 \tab 22 \tab 86 \tab 9 \tab 17 \tab 4 \tab 6 \tab 9 \tab 2 \tab 6 \tab 32 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}}
}
\concept{Family Mobility}
\keyword{variable}
