% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_163.R
\name{kdeduc3}
\alias{kdeduc3}
\title{Highest schl grade of r 3rd child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdeduc3
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
F3. Highest school grade third child completed.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1530/vshow}{GSS Data Explorer page for \code{kdeduc3}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 5 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2678 \tab 4 \tab 12 \tab 10 \tab 130 \tab 11 \tab 28 \tab 11 \tab 43 \tab 7 \tab 13 \tab 2 \tab 5 \tab 1 \tab 3 \tab 6 \tab 3 \tab 25 \tab 2992 \cr
   Total \tab 2678 \tab 4 \tab 12 \tab 10 \tab 130 \tab 11 \tab 28 \tab 11 \tab 43 \tab 7 \tab 13 \tab 2 \tab 5 \tab 1 \tab 3 \tab 6 \tab 3 \tab 25 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdpicked}},
\code{\link{kdwork1}},
\code{\link{kdwork2}}
}
\concept{Family Mobility}
\keyword{variable}
