% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_163.R
\name{kdeduc7}
\alias{kdeduc7}
\title{Highest schl grade of r 7th child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdeduc7
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
F7. Highest school grade seventh child completed.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1534/vshow}{GSS Data Explorer page for \code{kdeduc7}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 10 \tab 12 \tab 13 \tab 14 \tab 16 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1994 \tab 2964 \tab 4 \tab 10 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2992 \cr
   Total \tab 2964 \tab 4 \tab 10 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdpicked}},
\code{\link{kdwork1}},
\code{\link{kdwork2}}
}
\concept{Family Mobility}
\keyword{variable}
