% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_165.R
\name{kdindtag}
\alias{kdindtag}
\title{Kdind10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdindtag
}
\details{
Question KDIND10 Coding Tag
}
\section{Values}{

\itemize{
\item \verb{[1]} found
\item \verb{[2]} not found, assigned
\item \verb{[3]} not found, not assigned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6016/vshow}{GSS Data Explorer page for \code{kdindtag}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab found \tab not found, assigned \tab not found, not assigned \tab Total \cr
   1994 \tab - \tab 2938 \tab 52 \tab 2 \tab 2992 \cr
   Total \tab 0 \tab 2938 \tab 52 \tab 2 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdindstatus}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}}
}
\concept{Family Mobility}
\keyword{variable}
