% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_164.R
\name{kdocctag}
\alias{kdocctag}
\title{Kdocc10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdocctag
}
\details{
Question KDOCC10 Coding Tag
}
\section{Values}{

\itemize{
\item \verb{[1]} found
\item \verb{[2]} not found, assigned
\item \verb{[3]} not found, not assigned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6013/vshow}{GSS Data Explorer page for \code{kdocctag}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab found \tab not found, assigned \tab not found, not assigned \tab Total \cr
   1994 \tab - \tab 2938 \tab 51 \tab 3 \tab 2992 \cr
   Total \tab 0 \tab 2938 \tab 51 \tab 3 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
