% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_163.R
\name{kdpicked}
\alias{kdpicked}
\title{Number of child randomly chosen}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdpicked
}
\details{
Question NUMBER OF CHILD SELECTED FOR FOLLOW-UP QUESTIONS:
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1537/vshow}{GSS Data Explorer page for \code{kdpicked}}.
}

\seealso{
Other Family Mobility: 
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}}
}
\concept{Family Mobility}
\keyword{variable}
