% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_165.R
\name{kdpres105plus}
\alias{kdpres105plus}
\title{R child occupational prestige score using threshold method (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdpres105plus
}
\details{
Question Prestige of Respondent's Child's Occupation (2010)
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6018/vshow}{GSS Data Explorer page for \code{kdpres105plus}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 37 \tab 38 \tab 39 \tab 40 \tab 42 \tab 43 \tab 44 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 78 \tab 79 \tab 8 \tab 80 \tab 82 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 9 \tab 90 \tab 91 \tab 92 \tab 93 \tab 94 \tab 95 \tab 97 \tab Total \cr
   1994 \tab 2428 \tab 1 \tab 6 \tab 9 \tab 8 \tab 5 \tab 5 \tab 33 \tab 2 \tab 23 \tab 2 \tab 4 \tab 7 \tab 5 \tab 8 \tab 9 \tab 2 \tab 12 \tab 10 \tab 13 \tab 29 \tab 1 \tab 5 \tab 6 \tab 3 \tab 9 \tab 3 \tab 1 \tab 11 \tab 4 \tab 7 \tab 2 \tab 4 \tab 2 \tab 1 \tab 15 \tab 10 \tab 2 \tab 21 \tab 5 \tab 3 \tab 4 \tab 11 \tab 38 \tab 11 \tab 3 \tab 17 \tab 2 \tab 2 \tab 2 \tab 5 \tab 6 \tab 6 \tab 8 \tab 7 \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 4 \tab 8 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 7 \tab 21 \tab 3 \tab 18 \tab 2 \tab 2 \tab 6 \tab 6 \tab 8 \tab 1 \tab 3 \tab 10 \tab 3 \tab 2 \tab 2992 \cr
   Total \tab 2428 \tab 1 \tab 6 \tab 9 \tab 8 \tab 5 \tab 5 \tab 33 \tab 2 \tab 23 \tab 2 \tab 4 \tab 7 \tab 5 \tab 8 \tab 9 \tab 2 \tab 12 \tab 10 \tab 13 \tab 29 \tab 1 \tab 5 \tab 6 \tab 3 \tab 9 \tab 3 \tab 1 \tab 11 \tab 4 \tab 7 \tab 2 \tab 4 \tab 2 \tab 1 \tab 15 \tab 10 \tab 2 \tab 21 \tab 5 \tab 3 \tab 4 \tab 11 \tab 38 \tab 11 \tab 3 \tab 17 \tab 2 \tab 2 \tab 2 \tab 5 \tab 6 \tab 6 \tab 8 \tab 7 \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 4 \tab 8 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 7 \tab 21 \tab 3 \tab 18 \tab 2 \tab 2 \tab 6 \tab 6 \tab 8 \tab 1 \tab 3 \tab 10 \tab 3 \tab 2 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdpres10}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}}
}
\concept{Family Mobility}
\keyword{variable}
