% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_161.R
\name{kdrel1}
\alias{kdrel1}
\title{R relation to r 1st child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdrel1
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
D1. Relation to R. of first child
}
\section{Values}{

\itemize{
\item \verb{[1]} biological
\item \verb{[2]} adopted
\item \verb{[3]} step
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1510/vshow}{GSS Data Explorer page for \code{kdrel1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab biological \tab adopted \tab step \tab no answer \tab Total \cr
   1994 \tab 1860 \tab 1043 \tab 21 \tab 59 \tab 9 \tab 2992 \cr
   Total \tab 1860 \tab 1043 \tab 21 \tab 59 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdalive1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}}
}
\concept{Family Mobility}
\keyword{variable}
