% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_159.R
\name{kdsex1}
\alias{kdsex1}
\title{Sex of r 1st child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdsex1
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
A. What is (this/your next) child's first name?
B. Is (CHILD'S NAME) male or female?
C. In what year was (CHILD'S NAME) born?
D. Is (CHILD'S NAME) your biological child (B), adopted child (A), or step child (S)?
E. Is (CHILD'S NAME) alive?
F. (IF BORN BEFORE 1976) What is the highest grade or year of regular school that (CHILD'S NAME) ever completed?  CODE EXACT GRADE.
G. Do you have any other children?
B1. Sex of first child.
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1492/vshow}{GSS Data Explorer page for \code{kdsex1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1994 \tab 1860 \tab 593 \tab 530 \tab 9 \tab 2992 \cr
   Total \tab 1860 \tab 593 \tab 530 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{sphedpre105plus}},
\code{\link{spsibs}}
}
\concept{Family Mobility}
\keyword{variable}
