% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_159.R
\name{kdsex2}
\alias{kdsex2}
\title{Sex of r 2nd child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdsex2
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
B2. Sex of second child.
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1493/vshow}{GSS Data Explorer page for \code{kdsex2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 2105 \tab 439 \tab 438 \tab 10 \tab - \tab 2992 \cr
   Total \tab 2105 \tab 439 \tab 438 \tab 10 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{fstspyr}},
\code{\link{indfirst}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpicked}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{malive1}},
\code{\link{marnum}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{palive1}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind10}},
\code{\link{sbind80}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sbocc80}},
\code{\link{sboccindv}},
\code{\link{sboccstatus}},
\code{\link{sbocctag}},
\code{\link{sbpicked}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{spfam16}},
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre}},
\code{\link{sphedpre10}},
\code{\link{sphedpre105plus}},
\code{\link{sphedslf}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spsibs}},
\code{\link{spyrborn}},
\code{\link{wrkslfst}}

Other Children: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adoption}},
\code{\link{adoptkid}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{aidkids}},
\code{\link{allergic}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{cathsch}},
\code{\link{chembal}},
\code{\link{childs}},
\code{\link{chldben}},
\code{\link{chldcare}},
\code{\link{chldeduc}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldhrt}},
\code{\link{chldidel}},
\code{\link{chldidel_next}},
\code{\link{chldlove}},
\code{\link{chldmore}},
\code{\link{chldmorl}},
\code{\link{chldnum}},
\code{\link{chldpref_next}},
\code{\link{chldprob}},
\code{\link{chldsafe}},
\code{\link{chldsch}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldskls}},
\code{\link{chldsoon}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{chldvig}},
\code{\link{chrissch}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{comknows}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{discipln}},
\code{\link{eldersup}},
\code{\link{evkid}},
\code{\link{failure}},
\code{\link{famper4}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{fourkids}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{helpoth}},
\code{\link{hlth8}},
\code{\link{homesch}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{illphys}},
\code{\link{impkids}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpicked}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kid5up}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidhlth}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidlived}},
\code{\link{kidneedy}},
\code{\link{kidnofre}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidnum}},
\code{\link{kidout}},
\code{\link{kidpars}},
\code{\link{kidresp}},
\code{\link{kidskips}},
\code{\link{kidsocst}},
\code{\link{kidssol}},
\code{\link{kidtrble}},
\code{\link{kidxfilm}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{law7}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{mapaid}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{mntlill}},
\code{\link{morextrm}},
\code{\link{mykids}},
\code{\link{nextdoor}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obeythnk}},
\code{\link{obrespct}},
\code{\link{onekid}},
\code{\link{oth3}},
\code{\link{othkids}},
\code{\link{othsch}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{popular}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{raised}},
\code{\link{relsch}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{sellbaby}},
\code{\link{spanking}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spendeve}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}},
\code{\link{spwrkpar}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{stressfl}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{sufadult}},
\code{\link{thnkself}},
\code{\link{threkids}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{trbllaw}},
\code{\link{twokids}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{workhard}},
\code{\link{zombies}}

Other Household: 
\code{\link{adults}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{dinner}},
\code{\link{earnrs}},
\code{\link{fairhwrk}},
\code{\link{fambudgt}},
\code{\link{famfinan}},
\code{\link{famgen}},
\code{\link{famgen_7522}},
\code{\link{famgen_exp}},
\code{\link{famgen_exp_22}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{fewrksup}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{hefinfo}},
\code{\link{hhclean1}},
\code{\link{hhrace}},
\code{\link{hhtype}},
\code{\link{hhtype1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homeless}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{hubbywk1}},
\code{\link{hwduties}},
\code{\link{incdef}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{mrmom}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{ownhh}},
\code{\link{ownhh1}},
\code{\link{povline}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relhh1}},
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh14}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp11}},
\code{\link{relsp12}},
\code{\link{relsp13}},
\code{\link{relsp14}},
\code{\link{relsp2}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{respnum}},
\code{\link{rhhwork}},
\code{\link{rhmewrk}},
\code{\link{rplace}},
\code{\link{rplace_7522}},
\code{\link{rvisitor}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{sphhwork}},
\code{\link{sphmewrk}},
\code{\link{strsshme}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{timehome}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{visitors}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{womenben}},
\code{\link{womenhrt}}

Other Sex: 
\code{\link{acpttg}},
\code{\link{acqntsex}},
\code{\link{attractd}},
\code{\link{chngdocs}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{condom}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{family16sex}},
\code{\link{frndsex}},
\code{\link{harsexcl}},
\code{\link{harsexjb}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{kdsex1}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kidxfilm}},
\code{\link{knwclsex}},
\code{\link{matesex}},
\code{\link{numclerg}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{othersex}},
\code{\link{ownclerg}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{premars1}},
\code{\link{premarsx}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{relgendr}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sellsex}},
\code{\link{sex}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sexeduc}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexnow2}},
\code{\link{sexornt}},
\code{\link{sexsex}},
\code{\link{sexsex18}},
\code{\link{sexsex5}},
\code{\link{sexunion}},
\code{\link{spvtrfair}},
\code{\link{teensex}},
\code{\link{teensex1}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{xmarsex}},
\code{\link{xmarsex1}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Children}
\concept{Family Mobility}
\concept{Household}
\concept{Sex}
\keyword{variable}
