% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_159.R
\name{kdsex3}
\alias{kdsex3}
\title{Sex of r 3rd child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdsex3
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
B3. Sex of third child.
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1494/vshow}{GSS Data Explorer page for \code{kdsex3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1994 \tab 2500 \tab 239 \tab 242 \tab 11 \tab 2992 \cr
   Total \tab 2500 \tab 239 \tab 242 \tab 11 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{sphedpre105plus}},
\code{\link{spsibs}}
}
\concept{Family Mobility}
\keyword{variable}
