% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_292.R
\name{kdtmkid3}
\alias{kdtmkid3}
\title{Time child 3 spends with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdtmkid3
}
\details{
Question Would you say that (NAME OF CHILD) feels that the amount of time they spend with you is too much, about the right amount, or too little?
C. Third listed child?
}
\section{Values}{

\itemize{
\item \code{1} too much
\item \code{2} about the right amount
\item \code{3} too little
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2649/vshow}{GSS Data Explorer page for \code{kdtmkid3}}.
}

\seealso{
Other Information Society: 
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}}
}
\concept{Information Society}
\keyword{variable}
