% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_163.R
\name{kdwork2}
\alias{kdwork2}
\title{Does r child do paid or unpaid work?}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdwork2
}
\details{
Question Does (s/he) do any work at all, even if it's only a part-time job? Include unpaid work of 15 or more hours per week in a family business or farm.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1539/vshow}{GSS Data Explorer page for \code{kdwork2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2809 \tab 36 \tab 127 \tab 13 \tab 7 \tab 2992 \cr
   Total \tab 2809 \tab 36 \tab 127 \tab 13 \tab 7 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdpicked}},
\code{\link{kdwork1}}
}
\concept{Family Mobility}
\keyword{variable}
