% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_160.R
\name{kdyrbrn6}
\alias{kdyrbrn6}
\title{Birth year of r 6th child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdyrbrn6
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
C6. Birth year of sixth child.
}
\section{Values}{

\itemize{
\item \verb{[9996]} before 1976
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1506/vshow}{GSS Data Explorer page for \code{kdyrbrn6}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1928 \tab 1940 \tab 1945 \tab 1946 \tab 1949 \tab 1952 \tab 1953 \tab 1954 \tab 1955 \tab 1956 \tab 1957 \tab 1958 \tab 1959 \tab 1960 \tab 1961 \tab 1963 \tab 1964 \tab 1965 \tab 1966 \tab 1967 \tab 1968 \tab 1969 \tab 1970 \tab 1971 \tab 1973 \tab 1974 \tab 1975 \tab 1976 \tab 1983 \tab 1984 \tab 1985 \tab 1989 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2932 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2992 \cr
   Total \tab 2932 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}}
}
\concept{Family Mobility}
\keyword{variable}
