% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_160.R
\name{kdyrbrn7}
\alias{kdyrbrn7}
\title{Birth year of r 7th child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdyrbrn7
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
C7. Birth year of seventh child.
}
\section{Values}{

\itemize{
\item \code{9996} before 1976
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1507/vshow}{GSS Data Explorer page for \code{kdyrbrn7}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab 1931 \tab 1946 \tab 1951 \tab 1953 \tab 1955 \tab 1956 \tab 1957 \tab 1958 \tab 1959 \tab 1960 \tab 1961 \tab 1962 \tab 1963 \tab 1964 \tab 1966 \tab 1967 \tab 1971 \tab 1972 \tab 1976 \tab 1978 \tab 1980 \tab 1986 \tab 1988 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2955 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2992 \cr
   Total \tab 2955 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}}
}
\concept{Family Mobility}
\keyword{variable}
