% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_160.R
\name{kdyrbrn8}
\alias{kdyrbrn8}
\title{Birth year of r 8th child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdyrbrn8
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
C8. Birth year of eighth child.
}
\section{Values}{

\itemize{
\item \verb{[9996]} before 1976
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1508/vshow}{GSS Data Explorer page for \code{kdyrbrn8}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab 1932 \tab 1948 \tab 1954 \tab 1958 \tab 1959 \tab 1960 \tab 1961 \tab 1962 \tab 1964 \tab 1965 \tab 1971 \tab 1977 \tab 1990 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2975 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2992 \cr
   Total \tab 2975 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn9}}
}
\concept{Family Mobility}
\keyword{variable}
