% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{kid5up}
\alias{kid5up}
\title{Have children older than 5}
\source{
General Social Survey https://gss.norc.org
}
\description{
kid5up
}
\details{
Question Do you have any children older than 5?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1146/vshow}{GSS Data Explorer page for \code{kid5up}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 870 \tab 553 \tab 22 \tab 2832 \cr
   2000 \tab - \tab 1721 \tab 1093 \tab 3 \tab 2817 \cr
   Total \tab 1387 \tab 2591 \tab 1646 \tab 25 \tab 5649 \cr
}
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{givearts}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}},
\code{\link{relhrs2}}
}
\concept{Religion}
\keyword{variable}
