% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_457.R
\name{kidcall}
\alias{kidcall}
\title{How often does r contact by telephone or letter}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidcall
}
\details{
Question And how often do you have any other contact with this son or daughter besides visiting, either by telephone, letter, fax or e-mail?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} at least several times a week
\item \verb{[3]} at least once a week
\item \verb{[4]} at least once a month
\item \verb{[5]} several times a year
\item \verb{[6]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4191/vshow}{GSS Data Explorer page for \code{kidcall}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab daily \tab at least several times a week \tab at least once a week \tab at least once a month \tab several times a year \tab less often \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2397 \tab 101 \tab 97 \tab 88 \tab 33 \tab 26 \tab 14 \tab 1 \tab 8 \tab 2765 \cr
   Total \tab 2397 \tab 101 \tab 97 \tab 88 \tab 33 \tab 26 \tab 14 \tab 1 \tab 8 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{cousins}},
\code{\link{macall1}},
\code{\link{mavisit1}},
\code{\link{niecenep}},
\code{\link{pacall1}},
\code{\link{parslaw}},
\code{\link{pavisit1}},
\code{\link{sibinlaw}},
\code{\link{uncaunts}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
