% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_439.R
\name{kidedpar}
\alias{kidedpar}
\title{Govt intervene: child not sent to school}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidedpar
}
\details{
Question Sometimes public authorities intervene with parents in raising their children. Please indicate in each of the following cases how far you think authorities should go in dealing with a 10 year old child and his or her parents. Public authorities should...
G. The parents refuse to send their child to school because they wish to educate the child at home.
}
\section{Values}{

\itemize{
\item \verb{[1]} take no action
\item \verb{[2]} give warning
\item \verb{[3]} take child
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4034/vshow}{GSS Data Explorer page for \code{kidedpar}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab take no action \tab give warning \tab take child \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 140 \tab 373 \tab 82 \tab 67 \tab 15 \tab 1534 \cr
   Total \tab 857 \tab 140 \tab 373 \tab 82 \tab 67 \tab 15 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{beltup}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidout}},
\code{\link{kidskips}},
\code{\link{kidxfilm}},
\code{\link{nosmoke}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
