% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{kidempty}
\alias{kidempty}
\title{Kidless people lead empty lives}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidempty
}
\details{
Question Do you agree or disagree ...
F. People who have never had children lead empty lives.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4385/vshow}{GSS Data Explorer page for \code{kidempty}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 115 \tab 251 \tab 373 \tab 415 \tab 192 \tab 51 \tab 17 \tab 1481 \cr
   1994 \tab 1545 \tab 48 \tab 214 \tab 373 \tab 536 \tab 182 \tab 40 \tab 54 \tab 2992 \cr
   2002 \tab 1594 \tab 76 \tab 127 \tab 255 \tab 428 \tab 252 \tab 32 \tab 1 \tab 2765 \cr
   Total \tab 3206 \tab 239 \tab 592 \tab 1001 \tab 1379 \tab 626 \tab 123 \tab 72 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
