% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{kidjob}
\alias{kidjob}
\title{Children limit emp and career opp for one or both parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidjob
}
\details{
Question To what extent do you agree or disagree with the following statements?
D. Having children restricts the employment and career chances of one or both parents.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4460/vshow}{GSS Data Explorer page for \code{kidjob}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2012 \tab 672 \tab 41 \tab 381 \tab 218 \tab 492 \tab 126 \tab 38 \tab 6 \tab 1974 \cr
   Total \tab 672 \tab 41 \tab 381 \tab 218 \tab 492 \tab 126 \tab 38 \tab 6 \tab 1974 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
