% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{kidjoy}
\alias{kidjoy}
\title{Kids are life's greatest joy}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidjoy
}
\details{
Question Do you agree or disagree ...
B. Watching children grow up is life''s greatest joy.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4381/vshow}{GSS Data Explorer page for \code{kidjoy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 581 \tab 626 \tab 126 \tab 36 \tab 19 \tab 10 \tab 16 \tab 1481 \cr
   1994 \tab 1545 \tab 484 \tab 645 \tab 188 \tab 52 \tab 6 \tab 21 \tab 51 \tab 2992 \cr
   2002 \tab 1594 \tab 611 \tab 374 \tab 130 \tab 26 \tab 11 \tab 18 \tab 1 \tab 2765 \cr
   2012 \tab 672 \tab 550 \tab 558 \tab 120 \tab 40 \tab 3 \tab 26 \tab 5 \tab 1974 \cr
   Total \tab 3878 \tab 2226 \tab 2203 \tab 564 \tab 154 \tab 39 \tab 75 \tab 73 \tab 9212 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
