% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{kidless}
\alias{kidless}
\title{Kidless marriage is incomplete}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidless
}
\details{
Question Do you agree or disagree ...
D. A marriage without children is not fully complete.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4383/vshow}{GSS Data Explorer page for \code{kidless}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 196 \tab 427 \tab 324 \tab 305 \tab 102 \tab 44 \tab 16 \tab 1481 \cr
   Total \tab 67 \tab 196 \tab 427 \tab 324 \tab 305 \tab 102 \tab 44 \tab 16 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
