% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{kidmost}
\alias{kidmost}
\title{Do r have the most contact with son or daughter}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidmost
}
\details{
Question Of your children aged 18 and older, with whom do you have the most contact?
}
\section{Values}{

\itemize{
\item \code{1} with a son
\item \code{2} with a daughter
\item \code{3} i have no contact with any adult children
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4189/vshow}{GSS Data Explorer page for \code{kidmost}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
