% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{kidnofre}
\alias{kidnofre}
\title{Kids interfere with parents' freedom}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidnofre
}
\details{
Question Do you agree or disagree ...
C. Having children interferes too much with the freedom of parents.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4382/vshow}{GSS Data Explorer page for \code{kidnofre}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 25 \tab 130 \tab 186 \tab 656 \tab 370 \tab 25 \tab 22 \tab - \tab 1481 \cr
   1994 \tab 1545 \tab 14 \tab 105 \tab 216 \tab 749 \tab 288 \tab 18 \tab 57 \tab - \tab 2992 \cr
   2012 \tab 672 \tab 28 \tab 165 \tab 145 \tab 666 \tab 258 \tab 34 \tab 6 \tab - \tab 1974 \cr
   2022 \tab 2380 \tab 67 \tab 172 \tab 325 \tab 357 \tab 173 \tab 58 \tab 4 \tab 8 \tab 3544 \cr
   Total \tab 4664 \tab 134 \tab 572 \tab 872 \tab 2428 \tab 1089 \tab 135 \tab 89 \tab 8 \tab 9991 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidless}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
