% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{kidnow1}
\alias{kidnow1}
\title{R's children under 18 living with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidnow1
}
\details{
Question Are your biological or adopted children under age 18 living with you now at least some of the time?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, all of them live with me at least some of the time
\item \verb{[2]} some live with me at least some of the time
\item \verb{[3]} none live with me
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3917/vshow}{GSS Data Explorer page for \code{kidnow1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, all of them live with me at least some of the time \tab some live with me at least some of the time \tab none live with me \tab no answer \tab Total \cr
   2008 \tab 1942 \tab 19 \tab 2 \tab 3 \tab 57 \tab 2023 \cr
   Total \tab 1942 \tab 19 \tab 2 \tab 3 \tab 57 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidresp}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{toldsxor}}
}
\concept{Sexual Orientation}
\keyword{variable}
