% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{kidnow2}
\alias{kidnow2}
\title{R is living in a hh where children under 18 are living with r under r's care}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidnow2
}
\details{
Question Right now are you living in a household where any children under 18 are living with you at least some of the time and are under your care and responsibility?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3918/vshow}{GSS Data Explorer page for \code{kidnow2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1953 \tab 8 \tab 34 \tab 2 \tab 26 \tab 2023 \cr
   Total \tab 1953 \tab 8 \tab 34 \tab 2 \tab 26 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
