% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_439.R
\name{kidout}
\alias{kidout}
\title{Govt intervene: child stays out late}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidout
}
\details{
Question Sometimes public authorities intervene with parents in raising their children. Please indicate in each of the following cases how far you think authorities should go in dealing with a 10 year old child and his or her parents. Public authorities should...
C. The parents regularly let the child stay out late at night without knowing where the child is.
}
\section{Values}{

\itemize{
\item \code{1} take no action
\item \code{2} give warning
\item \code{3} take child
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4030/vshow}{GSS Data Explorer page for \code{kidout}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab take no action \tab give warning \tab take child \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 67 \tab 486 \tab 80 \tab 29 \tab 15 \tab 1534 \cr
   Total \tab 857 \tab 67 \tab 486 \tab 80 \tab 29 \tab 15 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{beltup}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidskips}},
\code{\link{kidxfilm}},
\code{\link{nosmoke}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
