% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{kidpars}
\alias{kidpars}
\title{Do adult children have duty to look after parent}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidpars
}
\details{
Question Please select a response to indicate how much you agree or disagree with each of the following statements.
A. Adult children have a duty to look after their elderly parents
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4229/vshow}{GSS Data Explorer page for \code{kidpars}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 406 \tab 417 \tab 198 \tab 79 \tab 25 \tab 17 \tab 7 \tab 2765 \cr
   2018 \tab 1175 \tab 371 \tab 458 \tab 234 \tab 78 \tab 18 \tab 11 \tab 3 \tab 2348 \cr
   Total \tab 2791 \tab 777 \tab 875 \tab 432 \tab 157 \tab 43 \tab 28 \tab 10 \tab 5113 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{helpjob}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{talkedto}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
