% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{kidsinhh}
\alias{kidsinhh}
\title{Does interviewer believe that there are children under the age of 15 present at}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidsinhh
}
\details{
Question Based on the features of this housing unit and the surrounding property/neighborhood that you can see, do you believe that there are children under the age of 15 present in the household?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6655/vshow}{GSS Data Explorer page for \code{kidsinhh}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab - \tab 831 \tab 1998 \tab 11 \tab 27 \tab - \tab 2867 \cr
   2018 \tab - \tab 688 \tab 1636 \tab 18 \tab 6 \tab - \tab 2348 \cr
   Total \tab 0 \tab 1519 \tab 3634 \tab 29 \tab 33 \tab 0 \tab 5215 \cr
}
}

\keyword{variable}
