% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{kidssolnv}
\alias{kidssolnv}
\title{R's kids living standard compared to r (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidssolnv
}
\details{
Question When your children are at the age you are now, do you think their standard of living will be much better, somewhat better, about the same, somewhat worse, or much worse than yours is now?
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} somewhat better
\item \verb{[3]} about the same
\item \verb{[4]} somewhat worse
\item \verb{[5]} much worse
\item \verb{[6]} no children (phone mode only; volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7619/vshow}{GSS Data Explorer page for \code{kidssolnv}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab much better \tab somewhat better \tab about the same \tab somewhat worse \tab much worse \tab no children (phone mode only; volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2719 \tab 261 \tab 364 \tab 331 \tab 222 \tab 75 \tab 10 \tab 6 \tab 1 \tab 43 \tab 4032 \cr
   2022 \tab 2954 \tab 111 \tab 158 \tab 135 \tab 110 \tab 51 \tab - \tab - \tab - \tab 25 \tab 3544 \cr
   Total \tab 5673 \tab 372 \tab 522 \tab 466 \tab 332 \tab 126 \tab 10 \tab 6 \tab 1 \tab 68 \tab 7576 \cr
}
}

\keyword{variable}
