% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{kidssolv}
\alias{kidssolv}
\title{R's kids living standard compared to r (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidssolv
}
\details{
Question When your children are at the age you are now, do you think their standard of living will be much better, somewhat better, about the same, somewhat worse, or much worse than yours is now?
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} somewhat better
\item \verb{[3]} about the same
\item \verb{[4]} somewhat worse
\item \verb{[5]} much worse
\item \verb{[6]} no children (phone mode only; volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7599/vshow}{GSS Data Explorer page for \code{kidssolv}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab much better \tab somewhat better \tab about the same \tab somewhat worse \tab much worse \tab no children (phone mode only; volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2673 \tab 262 \tab 359 \tab 278 \tab 185 \tab 77 \tab 189 \tab 3 \tab 2 \tab 4 \tab 4032 \cr
   2022 \tab 2966 \tab 111 \tab 134 \tab 110 \tab 94 \tab 47 \tab 81 \tab - \tab - \tab 1 \tab 3544 \cr
   2024 \tab 2714 \tab 74 \tab 141 \tab 108 \tab 116 \tab 72 \tab 80 \tab - \tab - \tab 4 \tab 3309 \cr
   Total \tab 8353 \tab 447 \tab 634 \tab 496 \tab 395 \tab 196 \tab 350 \tab 3 \tab 2 \tab 9 \tab 10885 \cr
}
}

\keyword{variable}
