% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_439.R
\name{kidxfilm}
\alias{kidxfilm}
\title{Govt intervene: chld allowed to see porn}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidxfilm
}
\details{
Question Sometimes public authorities intervene with parents in raising their children. Please indicate in each of the following cases how far you think authorities should go in dealing with a 10 year old child and his or her parents. Public authorities should...
H. The parents allow the child to watch violent or pornographic films.
}
\section{Values}{

\itemize{
\item \verb{[1]} take no action
\item \verb{[2]} give warning
\item \verb{[3]} take child
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4035/vshow}{GSS Data Explorer page for \code{kidxfilm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab take no action \tab give warning \tab take child \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1985 \tab 857 \tab 91 \tab 457 \tab 65 \tab 49 \tab 15 \tab - \tab 1534 \cr
   Total \tab 857 \tab 91 \tab 457 \tab 65 \tab 49 \tab 15 \tab 0 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidneedy}},
\code{\link{aidold}},
\code{\link{aidsmart}},
\code{\link{aidunemp}},
\code{\link{beltup}},
\code{\link{bribe}},
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{buspow}},
\code{\link{bustax}},
\code{\link{cctv}},
\code{\link{colop}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{cutdebt}},
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{databank}},
\code{\link{demworks}},
\code{\link{emonitor}},
\code{\link{eqincome}},
\code{\link{equalize}},
\code{\link{exploit}},
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{fewtrsty}},
\code{\link{gendereq}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{govtinfo}},
\code{\link{govtpow}},
\code{\link{grnlaws}},
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}},
\code{\link{hlphitec}},
\code{\link{hlpinflu}},
\code{\link{hlthcare}},
\code{\link{hsbasics}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{jobsall}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidout}},
\code{\link{kidskips}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{mustret}},
\code{\link{nosmoke}},
\code{\link{obeylaw}},
\code{\link{opfamily}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{polgbeco}},
\code{\link{polint}},
\code{\link{polsfair}},
\code{\link{prespop}},
\code{\link{pricecon}},
\code{\link{progtax}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{protstrs}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{prvdschl}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{racists}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{ratetone}},
\code{\link{revoltrs}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}},
\code{\link{unionsok}},
\code{\link{verdict}},
\code{\link{volacty2}},
\code{\link{volactyr}},
\code{\link{wotrial}},
\code{\link{youinflu}}

Other Children: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adoption}},
\code{\link{adoptkid}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{aidkids}},
\code{\link{allergic}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{cathsch}},
\code{\link{chembal}},
\code{\link{childs}},
\code{\link{chldben}},
\code{\link{chldcare}},
\code{\link{chldeduc}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldhrt}},
\code{\link{chldidel}},
\code{\link{chldidel_next}},
\code{\link{chldlove}},
\code{\link{chldmore}},
\code{\link{chldmorl}},
\code{\link{chldnum}},
\code{\link{chldpref_next}},
\code{\link{chldprob}},
\code{\link{chldsafe}},
\code{\link{chldsch}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldskls}},
\code{\link{chldsoon}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{chldvig}},
\code{\link{chrissch}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{comknows}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{discipln}},
\code{\link{eldersup}},
\code{\link{evkid}},
\code{\link{failure}},
\code{\link{famper4}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{fourkids}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{helpoth}},
\code{\link{hlth8}},
\code{\link{homesch}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{illphys}},
\code{\link{impkids}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpicked}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kid5up}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidhlth}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidlived}},
\code{\link{kidneedy}},
\code{\link{kidnofre}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidnum}},
\code{\link{kidout}},
\code{\link{kidpars}},
\code{\link{kidresp}},
\code{\link{kidskips}},
\code{\link{kidsocst}},
\code{\link{kidssol}},
\code{\link{kidtrble}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{law7}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{mapaid}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{mntlill}},
\code{\link{morextrm}},
\code{\link{mykids}},
\code{\link{nextdoor}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obeythnk}},
\code{\link{obrespct}},
\code{\link{onekid}},
\code{\link{oth3}},
\code{\link{othkids}},
\code{\link{othsch}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{popular}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{raised}},
\code{\link{relsch}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{sellbaby}},
\code{\link{spanking}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spendeve}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}},
\code{\link{spwrkpar}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{stressfl}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{sufadult}},
\code{\link{thnkself}},
\code{\link{threkids}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{trbllaw}},
\code{\link{twokids}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{workhard}},
\code{\link{zombies}}

Other Movies: 
\code{\link{inffilms}},
\code{\link{seemovie}},
\code{\link{xmovie}},
\code{\link{xmovie1}}

Other Sex: 
\code{\link{acpttg}},
\code{\link{acqntsex}},
\code{\link{attractd}},
\code{\link{chngdocs}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{condom}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{family16sex}},
\code{\link{frndsex}},
\code{\link{harsexcl}},
\code{\link{harsexjb}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{knwclsex}},
\code{\link{matesex}},
\code{\link{numclerg}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{othersex}},
\code{\link{ownclerg}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{premars1}},
\code{\link{premarsx}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{relgendr}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sellsex}},
\code{\link{sex}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sexeduc}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexnow2}},
\code{\link{sexornt}},
\code{\link{sexsex}},
\code{\link{sexsex18}},
\code{\link{sexsex5}},
\code{\link{sexunion}},
\code{\link{spvtrfair}},
\code{\link{teensex}},
\code{\link{teensex1}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{xmarsex}},
\code{\link{xmarsex1}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Children}
\concept{ISSP Role of Gov't}
\concept{Movies}
\concept{Sex}
\keyword{variable}
