% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{kindpers}
\alias{kindpers}
\title{R is a kind person}
\source{
General Social Survey https://gss.norc.org
}
\description{
kindpers
}
\details{
Question For each of the following, please indicate how well the description applies to you by circling one number:
D. A kind person. Is this ……
}
\section{Values}{

\itemize{
\item \code{1} a very good description
\item \code{2} a good description
\item \code{3} a fair description
\item \code{4} not a very good description
\item \code{5} not a good description at all
\item \code{9} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3050/vshow}{GSS Data Explorer page for \code{kindpers}}.
}

\seealso{
Other Genes and the Environment: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{intrwght}},
\code{\link{moregood}},
\code{\link{owndoing}},
\code{\link{sadblue}},
\code{\link{selfish}}
}
\concept{Genes and the Environment}
\keyword{variable}
