% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{knomentl}
\alias{knomentl}
\title{Gov should have the right to know}
\source{
General Social Survey https://gss.norc.org
}
\description{
knomentl
}
\details{
Question Please tell me which one of the statements on the card best describes what the government has the right to know. Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to know:
}
\section{Values}{

\itemize{
\item \verb{[1]} nothing about mental health
\item \verb{[2]} whether currently consulting
\item \verb{[3]} ever consulted a mental
\item \verb{[4]} ever consulted & diagnosis
\item \verb{[5]} diagnosis & confident. info
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5226/vshow}{GSS Data Explorer page for \code{knomentl}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab nothing about mental health \tab whether currently consulting \tab ever consulted a mental \tab ever consulted & diagnosis \tab diagnosis & confident. info \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1444 \tab 95 \tab 174 \tab 122 \tab 367 \tab 582 \tab 86 \tab 34 \tab - \tab 2904 \cr
   1998 \tab 1445 \tab 66 \tab 160 \tab 136 \tab 381 \tab 522 \tab 105 \tab 17 \tab - \tab 2832 \cr
   2000 \tab 1398 \tab 84 \tab 135 \tab 143 \tab 336 \tab 596 \tab 98 \tab 27 \tab - \tab 2817 \cr
   Total \tab 4287 \tab 245 \tab 469 \tab 401 \tab 1084 \tab 1700 \tab 289 \tab 78 \tab 0 \tab 8553 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askcrime}},
\code{\link{askdrink}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{askmentl}},
\code{\link{asksexor}},
\code{\link{bugging}},
\code{\link{camwrk}},
\code{\link{chkfinan}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{chktravl}},
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{compfin}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{leakinfo}},
\code{\link{lietest}},
\code{\link{nuclrwar}},
\code{\link{punarms}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}},
\code{\link{rptcowrk}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secdocs}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{sectech}},
\code{\link{secterr}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{srchwrk}},
\code{\link{takearms}},
\code{\link{taketrck}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{testdrug}},
\code{\link{usspy}},
\code{\link{usterror}}

Other Federal Government: 
\code{\link{advfront}},
\code{\link{affrmact}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{concong}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{congeth}},
\code{\link{congovt}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{demworks}},
\code{\link{equalize}},
\code{\link{fedtrust}},
\code{\link{govcheat}},
\code{\link{govtpow}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hlthcare}},
\code{\link{hspgovt}},
\code{\link{immcosts}},
\code{\link{immfare}},
\code{\link{jobaff}},
\code{\link{jobsall}},
\code{\link{leadsci}},
\code{\link{mustret}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdiscm}},
\code{\link{natdisin}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natrecon}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{owneth}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{pricecon}},
\code{\link{reqinfo}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{scresrch}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secterr}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{spmentl}},
\code{\link{spnatdis}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{taxcheat}},
\code{\link{taxspend}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{whtgovt}}
}
\concept{Federal Government}
\concept{National Security}
\keyword{variable}
