% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_292.R
\name{knowkid1}
\alias{knowkid1}
\title{R knows what child is doing on weekend}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowkid1
}
\details{
Question Thinking now only about weekends (SATURDAY AND SUNDAY) and only when your child is awake–would you say that you know exactly what (NAME OF CHILD) is doing on the weekend almost all of the time, most of the time, about half of the time, some of the time or hardly any of the time?
A. First listed child?
}
\section{Values}{

\itemize{
\item \code{1} almost all of the time
\item \code{2} most of the time
\item \code{3} about half of the time
\item \code{4} some of the time
\item \code{5} hardly any of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2651/vshow}{GSS Data Explorer page for \code{knowkid1}}.
}

\seealso{
Other Information Society: 
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid2}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}}

Other Children: 
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid2}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}}
}
\concept{Children}
\concept{Information Society}
\keyword{variable}
