% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_102.R
\name{known4}
\alias{known4}
\title{How long has r known person #4?}
\source{
General Social Survey https://gss.norc.org
}
\description{
known4
}
\details{
Question Have you known (NAME) for less than 3 years, 3 to 6 years, or more than 6 years?  ASK FOR EACH NAME.
4. Known Fourth Person
}
\section{Values}{

\itemize{
\item \verb{[1]} lt 3 years
\item \verb{[2]} 3-6 years
\item \verb{[3]} 6 years +
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/938/vshow}{GSS Data Explorer page for \code{known4}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab lt 3 years \tab 3-6 years \tab 6 years + \tab don't know \tab no answer \tab Total \cr
   1985 \tab 920 \tab 73 \tab 105 \tab 427 \tab 1 \tab 8 \tab 1534 \cr
   Total \tab 920 \tab 73 \tab 105 \tab 427 \tab 1 \tab 8 \tab 1534 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}}
}
\concept{Social Networks}
\keyword{variable}
