% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{knowschd}
\alias{knowschd}
\title{How far in advance know work schedule}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowschd
}
\details{
Question How far in advance do you usually know what days and hours you will need to work?
}
\section{Values}{

\itemize{
\item \verb{[1]} one week or less
\item \verb{[2]} between 1 and 2 weeks
\item \verb{[3]} between 3 and 4 weeks
\item \verb{[4]} 4 weeks or more
\item \verb{[5]} my schedule never changes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2839/vshow}{GSS Data Explorer page for \code{knowschd}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab one week or less \tab between 1 and 2 weeks \tab between 3 and 4 weeks \tab 4 weeks or more \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2014 \tab 1288 \tab 540 \tab 165 \tab 65 \tab 459 \tab 13 \tab 8 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 599 \tab 183 \tab 85 \tab 531 \tab 20 \tab 30 \tab - \tab 2348 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 2188 \tab 1139 \tab 348 \tab 150 \tab 990 \tab 33 \tab 38 \tab 3544 \tab 8430 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{satjob1}},
\code{\link{slpprblm}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}

Other Job: 
\code{\link{satjob1}},
\code{\link{stress12}},
\code{\link{usetech}}
}
\concept{Employment}
\concept{Job}
\concept{Quality of Working Life}
\keyword{variable}
