% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{knowsol}
\alias{knowsol}
\title{Know solutions to environment issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowsol
}
\details{
Question F2. And how much do you feel you know about solutions to these sorts of environmental problems?  Please tell me what you think, where 1 indicates you feel you know nothing at all and 5 indicates you feel you know a great deal.
}
\section{Values}{

\itemize{
\item \code{1} know nothing at all
\item \code{5} know a great deal
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4806/vshow}{GSS Data Explorer page for \code{knowsol}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab know nothing at all \tab 2 \tab 3 \tab 4 \tab know a great deal \tab don't know \tab no answer \tab Total \cr
   2010 \tab 614 \tab 262 \tab 380 \tab 538 \tab 157 \tab 41 \tab 44 \tab 8 \tab 2044 \cr
   Total \tab 614 \tab 262 \tab 380 \tab 538 \tab 157 \tab 41 \tab 44 \tab 8 \tab 2044 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
