% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{knowwhat}
\alias{knowwhat}
\title{R knows what's expected on job}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowwhat
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
G. On my job, I know exactly what is expected of me
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2780/vshow}{GSS Data Explorer page for \code{knowwhat}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 772 \tab 890 \tab 91 \tab 22 \tab 8 \tab 13 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 721 \tab 862 \tab 111 \tab 19 \tab 9 \tab 12 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 468 \tab 626 \tab 58 \tab 11 \tab 1 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 516 \tab 670 \tab 53 \tab 4 \tab 1 \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 514 \tab 832 \tab 60 \tab 7 \tab 4 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 752 \tab 1052 \tab 149 \tab 13 \tab 5 \tab 8 \tab 19 \tab 3544 \cr
   Total \tab 8336 \tab 3743 \tab 4932 \tab 522 \tab 76 \tab 28 \tab 93 \tab 19 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
