% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{knwcause}
\alias{knwcause}
\title{Know causes of environment issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwcause
}
\details{
Question F1. How much do you feel you know about the causes of these sorts of environmental problems? Please tell me what you think, where 1 indicates you feel you know nothing at all and 5 indicates you feel you know a great deal.
}
\section{Values}{

\itemize{
\item \verb{[1]} know nothing at all
\item \verb{[5]} know a great deal
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4805/vshow}{GSS Data Explorer page for \code{knwcause}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab know nothing at all \tab 2 \tab 3 \tab 4 \tab know a great deal \tab don't know \tab no answer \tab Total \cr
   2010 \tab 614 \tab 168 \tab 229 \tab 599 \tab 284 \tab 100 \tab 43 \tab 7 \tab 2044 \cr
   Total \tab 614 \tab 168 \tab 229 \tab 599 \tab 284 \tab 100 \tab 43 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
