% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_432.R
\name{knwclsex}
\alias{knwclsex}
\title{R knows other people who were obj of sexual advances}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwclsex
}
\details{
Question Do you know of other people who, as adults, were the object of sexual advances or propositions from a minister, priest, rabbi, or other clergyperson or religious leader in a place you were attending religious services?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3966/vshow}{GSS Data Explorer page for \code{knwclsex}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 233 \tab 162 \tab 1600 \tab 3 \tab 25 \tab 2023 \cr
   Total \tab 233 \tab 162 \tab 1600 \tab 3 \tab 25 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clsclsex}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
