% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{knwecon}
\alias{knwecon}
\title{Level of r's information: economic policy}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwecon
}
\details{
Question We want to ask about how much information you have on various topics. Please look at Card B21. For each of the following areas, please indicate whether you are very informed, somewhat informed, neither informed nor uninformed, somewhat uniformed, or very uninformed about the issues.
B. Economic policy
}
\section{Values}{

\itemize{
\item \verb{[1]} very informed
\item \verb{[2]} somewhat informed
\item \verb{[3]} neither informed nor uninformed
\item \verb{[4]} somewhat uninformed
\item \verb{[5]} very uninformed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3415/vshow}{GSS Data Explorer page for \code{knwecon}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very informed \tab somewhat informed \tab neither informed nor uninformed \tab somewhat uninformed \tab very uninformed \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 93 \tab 804 \tab 413 \tab 299 \tab 223 \tab 27 \tab 5 \tab 4510 \cr
   Total \tab 2646 \tab 93 \tab 804 \tab 413 \tab 299 \tab 223 \tab 27 \tab 5 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{biosci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{medsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
