% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{knwmhosp}
\alias{knwmhosp}
\title{Know anyone who was in a hospital}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwmhosp
}
\details{
Question Did you ever know anyone who was in a hospital because of a mental illness?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2058/vshow}{GSS Data Explorer page for \code{knwmhosp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 2168 \tab 366 \tab 355 \tab 8 \tab 7 \tab - \tab - \tab 2904 \cr
   2002 \tab 1859 \tab 561 \tab 343 \tab 2 \tab - \tab - \tab - \tab 2765 \cr
   2024 \tab 1664 \tab 942 \tab 673 \tab 9 \tab 4 \tab - \tab 17 \tab 3309 \cr
   Total \tab 5691 \tab 1869 \tab 1371 \tab 19 \tab 11 \tab 0 \tab 17 \tab 8978 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}}

Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}}
}
\concept{Mental Health}
\keyword{variable}
