% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_221.R
\name{knwmw3}
\alias{knwmw3}
\title{Gender identity of person 3}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwmw3
}
\details{
Question Is (NAME 3), the third person, a man or a woman?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} transgender ftm
\item \code{4} transgender mtf
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7097/vshow}{GSS Data Explorer page for \code{knwmw3}}.
}

\keyword{variable}
