% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_221.R
\name{knwmw4}
\alias{knwmw4}
\title{Gender identity of person 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwmw4
}
\details{
Question Is (NAME 4), the fourth person, a man or a woman?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[3]} transgender ftm
\item \verb{[4]} transgender mtf
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7098/vshow}{GSS Data Explorer page for \code{knwmw4}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab male \tab female \tab other \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2190 \tab 69 \tab 80 \tab 2 \tab 4 \tab 3 \tab 2348 \cr
   Total \tab 2190 \tab 69 \tab 80 \tab 2 \tab 4 \tab 3 \tab 2348 \cr
}
}

\keyword{variable}
