% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{knwnproz}
\alias{knwnproz}
\title{Have you known anyone who took prozac}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwnproz
}
\details{
Question Next, I''d like to ask you some questions about the medication called Prozac.
C. Have you ever personally known anyone who took Prozac?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2222/vshow}{GSS Data Explorer page for \code{knwnproz}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 670 \tab 470 \tab 41 \tab 16 \tab 2832 \cr
   2006 \tab 3238 \tab 735 \tab 502 \tab 20 \tab 15 \tab 4510 \cr
   Total \tab 4873 \tab 1405 \tab 972 \tab 61 \tab 31 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
