% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{knwpatnt}
\alias{knwpatnt}
\title{Known anyone seeing a psychologist etc?}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwpatnt
}
\details{
Question Have you ever known anyone (other than persons mentioned in KNWMHOSP) who was seeing a psychologist, mental health professional, social worker or other counselor?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2065/vshow}{GSS Data Explorer page for \code{knwpatnt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2168 \tab 388 \tab 271 \tab 11 \tab 66 \tab 2904 \cr
   2002 \tab 1859 \tab 728 \tab 176 \tab 2 \tab - \tab 2765 \cr
   Total \tab 4027 \tab 1116 \tab 447 \tab 13 \tab 66 \tab 5669 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo1}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}}
}
\concept{Mental Health}
\keyword{variable}
