% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{knwpolar}
\alias{knwpolar}
\title{Level of r's information: the north and south poles}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwpolar
}
\details{
Question We want to ask about how much information you have on various topics. Please look at Card B21. For each of the following areas, please indicate whether you are very informed, somewhat informed, neither informed nor uninformed, somewhat uniformed, or very uninformed about the issues.
E. The North and South poles
}
\section{Values}{

\itemize{
\item \verb{[1]} very informed
\item \verb{[2]} somewhat informed
\item \verb{[3]} neither informed nor uninformed
\item \verb{[4]} somewhat uninformed
\item \verb{[5]} very uninformed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3418/vshow}{GSS Data Explorer page for \code{knwpolar}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very informed \tab somewhat informed \tab neither informed nor uninformed \tab somewhat uninformed \tab very uninformed \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 148 \tab 625 \tab 480 \tab 292 \tab 290 \tab 27 \tab 2 \tab 4510 \cr
   2010 \tab 1338 \tab 61 \tab 305 \tab 147 \tab 90 \tab 92 \tab 4 \tab 7 \tab 2044 \cr
   Total \tab 3984 \tab 209 \tab 930 \tab 627 \tab 382 \tab 382 \tab 31 \tab 9 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{biosci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwsci}},
\code{\link{medsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
