% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{knwsexor}
\alias{knwsexor}
\title{R knows anyone who is gay, lesbian, or bisexual}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwsexor
}
\details{
Question Among your immediate family members, relatives, neighbors, co-workers, or close friends, are any of them gay, lesbian, or bisexual as far as you know?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8293/vshow}{GSS Data Explorer page for \code{knwsexor}}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab yes \tab no \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 721 \tab 340 \tab 2244 \tab 4 \tab 3309 \cr
   Total \tab 0 \tab 721 \tab 340 \tab 2244 \tab 4 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{happy_next}},
\code{\link{perla_self}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{happy_next}},
\code{\link{perla_self}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}
}
\concept{GSS Next}
\keyword{variable}
