% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_257.R
\name{knwwht}
\alias{knwwht}
\title{Does r personally any whites}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwwht
}
\details{
Question Do you personally know any...?
A. Whites?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2330/vshow}{GSS Data Explorer page for \code{knwwht}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2462 \tab 308 \tab 23 \tab 24 \tab 2817 \cr
   Total \tab 2462 \tab 308 \tab 23 \tab 24 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{ethchng}},
\code{\link{hispwork}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
