% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{laborpow}
\alias{laborpow}
\title{Power of labor unions in r's country}
\source{
General Social Survey https://gss.norc.org
}
\description{
laborpow
}
\details{
Question Do you think that labor unions in this country have too much power or too little power?
}
\section{Values}{

\itemize{
\item \code{1} far too much power
\item \code{2} too much power
\item \code{3} about the right amount of power
\item \code{4} too little power
\item \code{5} far too little power
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1996 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4068/vshow}{GSS Data Explorer page for \code{laborpow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab far too much power \tab too much power \tab about the right amount of power \tab too little power \tab far too little power \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 157 \tab 241 \tab 176 \tab 42 \tab 7 \tab 43 \tab 11 \tab 1534 \cr
   1990 \tab 155 \tab 170 \tab 345 \tab 413 \tab 117 \tab 24 \tab 126 \tab 22 \tab 1372 \cr
   1996 \tab 1572 \tab 187 \tab 327 \tab 455 \tab 152 \tab 30 \tab 154 \tab 27 \tab 2904 \cr
   Total \tab 2584 \tab 514 \tab 913 \tab 1044 \tab 311 \tab 61 \tab 323 \tab 60 \tab 5810 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
