% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{laidoff}
\alias{laidoff}
\title{R was laid off main job last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
laidoff
}
\details{
Question Were you laid off your main job at any time in the last year?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2818/vshow}{GSS Data Explorer page for \code{laidoff}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 969 \tab 131 \tab 1644 \tab 6 \tab 15 \tab - \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 115 \tab 1602 \tab 3 \tab 14 \tab - \tab - \tab 4510 \cr
   2010 \tab 857 \tab 112 \tab 1048 \tab 3 \tab 24 \tab - \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 69 \tab 1175 \tab - \tab 6 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 67 \tab 1346 \tab 4 \tab 31 \tab - \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 115 \tab 1839 \tab 7 \tab 14 \tab 14 \tab - \tab 3544 \cr
   Total \tab 8345 \tab 609 \tab 8654 \tab 23 \tab 104 \tab 14 \tab 0 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{chngtme}},
\code{\link{condemnd}},
\code{\link{cowrkhlp}},
\code{\link{cowrkint}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{fairearn}},
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{fringeok}},
\code{\link{handmove}},
\code{\link{haveinfo}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{hlpequip}},
\code{\link{hlthdays}},
\code{\link{hrsrelax}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{hyperten}},
\code{\link{jobfind1}},
\code{\link{jobsecok}},
\code{\link{knowschd}},
\code{\link{knowschd1}},
\code{\link{knowwhat}},
\code{\link{learnnew}},
\code{\link{lotofsay}},
\code{\link{manvsemp}},
\code{\link{misswork}},
\code{\link{mntlhlth}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{myskills}},
\code{\link{ntwkhard}},
\code{\link{opdevel}},
\code{\link{overwork}},
\code{\link{painarms}},
\code{\link{partteam}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{prodctiv}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{respect}},
\code{\link{rincblls}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{satjob1}},
\code{\link{secondwk}},
\code{\link{setthngs}},
\code{\link{slpprblm}},
\code{\link{spvtrfair}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{suphelp}},
\code{\link{teamsafe}},
\code{\link{toofewwk}},
\code{\link{trainops}},
\code{\link{trdunion}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{usedup}},
\code{\link{usetech}},
\code{\link{waypaid}},
\code{\link{weight}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbonus}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkpraise}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksmooth}},
\code{\link{wktopsat}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrktime}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Employment: 
\code{\link{accntsci}},
\code{\link{advances}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compwage}},
\code{\link{curempyr}},
\code{\link{denyrais}},
\code{\link{dowell}},
\code{\link{drugtest}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empyears}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{evunemp}},
\code{\link{expernc}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findjob}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{futpromo}},
\code{\link{harjob5}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{ignorwk}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobyears}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{knowschd}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leavejb}},
\code{\link{liedcwkr}},
\code{\link{losejb12}},
\code{\link{losejob5}},
\code{\link{lostjobs}},
\code{\link{ltrsref}},
\code{\link{marrcoun}},
\code{\link{medtreat}},
\code{\link{moredays}},
\code{\link{moretrde}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{neverwk}},
\code{\link{newjobs}},
\code{\link{numemp}},
\code{\link{numemply}},
\code{\link{numemps}},
\code{\link{numemps5}},
\code{\link{numemps_0422}},
\code{\link{numown}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{ownbiz}},
\code{\link{paidemps}},
\code{\link{partfull}},
\code{\link{paysame}},
\code{\link{physical}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{ryrsfirm}},
\code{\link{samejob}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{skiltest}},
\code{\link{slsmnshp}},
\code{\link{spwrkgvt}},
\code{\link{startbiz}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totunemp}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{whywkhme}},
\code{\link{wkbhvrs}},
\code{\link{work3yrs}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrkstat}},
\code{\link{wrktype}},
\code{\link{wrkyears}},
\code{\link{yearbiz}},
\code{\link{yearsjob}},
\code{\link{yrsfirm}}
}
\concept{Employment}
\concept{Quality of Working Life}
\keyword{variable}
