% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_254.R
\name{langcom}
\alias{langcom}
\title{How often are other languages heard in community}
\source{
General Social Survey https://gss.norc.org
}
\description{
langcom
}
\details{
Question A. How often do you hear languages other than English spoken in the local community where you live?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a week
\item \code{3} once a week
\item \code{4} several times a week
\item \code{5} once a day
\item \code{6} several times a day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2292/vshow}{GSS Data Explorer page for \code{langcom}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab less than once a week \tab once a week \tab several times a week \tab once a day \tab several times a day \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 380 \tab 251 \tab 142 \tab 251 \tab 90 \tab 258 \tab 13 \tab 13 \tab 2817 \cr
   Total \tab 1419 \tab 380 \tab 251 \tab 142 \tab 251 \tab 90 \tab 258 \tab 13 \tab 13 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang1}},
\code{\link{granlang}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{twolang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
