% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_116.R
\name{lapsed}
\alias{lapsed}
\title{Did r ever lapse in church attendance}
\source{
General Social Survey https://gss.norc.org
}
\description{
lapsed
}
\details{
Question Since you were 16 years old, has there ever been a period of 2 years or more when you did not attend religious services?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1072/vshow}{GSS Data Explorer page for \code{lapsed}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1988 \tab - \tab 807 \tab 673 \tab 1 \tab 1481 \cr
   Total \tab 0 \tab 807 \tab 673 \tab 1 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{god}},
\code{\link{grace16}},
\code{\link{join16}},
\code{\link{reborn}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{tvrelig}}
}
\concept{Religion}
\keyword{variable}
