% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{lastwkmo}
\alias{lastwkmo}
\title{What month were you last working}
\source{
General Social Survey https://gss.norc.org
}
\description{
lastwkmo
}
\details{
Question When were you last working for pay, that is what month and year were you last working?
\enumerate{
\item MONTH
}
}
\section{Values}{

\itemize{
\item \code{1} january
\item \code{2} february
\item \code{3} march
\item \code{4} april
\item \code{5} may
\item \code{6} june
\item \code{7} july
\item \code{8} august
\item \code{9} september
\item \code{10} october
\item \code{11} november
\item \code{12} december
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2868/vshow}{GSS Data Explorer page for \code{lastwkmo}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{cowrkhrd}},
\code{\link{donothng}},
\code{\link{lastwkyr}},
\code{\link{talkemp}},
\code{\link{talksup}}
}
\concept{Employee Compensation}
\keyword{variable}
