% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{lastyrwk}
\alias{lastyrwk}
\title{Thinking about last full-time job, what yr r worked last}
\source{
General Social Survey https://gss.norc.org
}
\description{
lastyrwk
}
\details{
Question Thinking about your last full-time job, what year did you last work?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3893/vshow}{GSS Data Explorer page for \code{lastyrwk}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab 1974 \tab 1976 \tab 1980 \tab 1986 \tab 1989 \tab 1991 \tab 1995 \tab 1997 \tab 1999 \tab 2000 \tab 2001 \tab 2002 \tab 2003 \tab 2004 \tab 2005 \tab 2006 \tab 2007 \tab 2008 \tab don't know \tab Total \cr
   2008 \tab 1797 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 2 \tab 20 \tab 35 \tab 67 \tab 79 \tab 3 \tab 2023 \cr
   Total \tab 1797 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 2 \tab 20 \tab 35 \tab 67 \tab 79 \tab 3 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
