% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{laundry1}
\alias{laundry1}
\title{Who in household does laundry}
\source{
General Social Survey https://gss.norc.org
}
\description{
laundry1
}
\details{
Question In your household who does the following things...?
A. Does the laundry?
}
\section{Values}{

\itemize{
\item \verb{[1]} always me
\item \verb{[2]} usually me
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually my spouse/partner
\item \verb{[5]} always my spouse/partner
\item \verb{[6]} is done by a third person
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4429/vshow}{GSS Data Explorer page for \code{laundry1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab always me \tab usually me \tab about equal or both together \tab usually my spouse/partner \tab always my spouse/partner \tab is done by a third person \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 2058 \tab 170 \tab 140 \tab 187 \tab 151 \tab 22 \tab 10 \tab 6 \tab 21 \tab - \tab 2765 \cr
   2012 \tab 1274 \tab 187 \tab 109 \tab 176 \tab 123 \tab 87 \tab 14 \tab 2 \tab 2 \tab - \tab 1974 \cr
   2022 \tab 3024 \tab 91 \tab 68 \tab 184 \tab 97 \tab 66 \tab 6 \tab 6 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 6356 \tab 448 \tab 317 \tab 547 \tab 371 \tab 175 \tab 30 \tab 14 \tab 24 \tab 1 \tab 8283 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{repairs1}},
\code{\link{shopfood}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
