% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{lawimp}
\alias{lawimp}
\title{Importance of law enforcement issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawimp
}
\details{
Question How important is the law enforcement issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most imp
\item \verb{[2]} important
\item \verb{[3]} not very imp
\item \verb{[4]} not important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/281/vshow}{GSS Data Explorer page for \code{lawimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not important \tab don't know \tab no answer \tab Total \cr
   1984 \tab 484 \tab 349 \tab 568 \tab 59 \tab 6 \tab 5 \tab 2 \tab 1473 \cr
   Total \tab 484 \tab 349 \tab 568 \tab 59 \tab 6 \tab 5 \tab 2 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{lawfirm}},
\code{\link{lawinfo}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}},
\code{\link{wirtap}}
}
\concept{Core}
\keyword{variable}
