% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{lawyer1}
\alias{lawyer1}
\title{What was the most important reason?}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawyer1
}
\details{
Question Think about the most recent time you hired a lawyer. What made you choose this law firm or attorney? (Choose all that apply)
H. MOST IMPORTANT REASON FROM ABOVE.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1836/vshow}{GSS Data Explorer page for \code{lawyer1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2736 \tab 40 \tab 51 \tab 17 \tab 33 \tab 4 \tab 10 \tab 3 \tab 1 \tab 9 \tab 2904 \cr
   Total \tab 2736 \tab 40 \tab 51 \tab 17 \tab 33 \tab 4 \tab 10 \tab 3 \tab 1 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{rellaw}},
\code{\link{satlawyr}}
}
\concept{Markets}
\keyword{variable}
