% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{lawyerd}
\alias{lawyerd}
\title{General reputation in the community}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawyerd
}
\details{
Question Think about the most recent time you hired a lawyer. What made you choose this law firm or attorney? (Choose all that apply)
D. General reputation in the community.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1832/vshow}{GSS Data Explorer page for \code{lawyerd}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2254 \tab 132 \tab 391 \tab 127 \tab 2904 \cr
   Total \tab 2254 \tab 132 \tab 391 \tab 127 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{rellaw}},
\code{\link{satlawyr}}
}
\concept{Markets}
\keyword{variable}
