% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{lawyerf}
\alias{lawyerf}
\title{I searched for the best price.}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawyerf
}
\details{
Question Think about the most recent time you hired a lawyer. What made you choose this law firm or attorney? (Choose all that apply)
F. I searched for the best price.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1834/vshow}{GSS Data Explorer page for \code{lawyerf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2254 \tab 45 \tab 453 \tab 1 \tab 151 \tab 2904 \cr
   Total \tab 2254 \tab 45 \tab 453 \tab 1 \tab 151 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerg}},
\code{\link{rellaw}},
\code{\link{satlawyr}}
}
\concept{Markets}
\keyword{variable}
