% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_200.R
\name{layoffs}
\alias{layoffs}
\title{Who has the greatest right to keep their jobs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
layoffs
}
\details{
Question From time to time companies must lay-off some of their employees. If some have to be laid-off, who, in your opinion, has the greatest right to keep their jobs?
}
\section{Values}{

\itemize{
\item \verb{[1]} those who have worked longest
\item \verb{[2]} those who have the greatest needs
\item \verb{[3]} those who are the best workers
\item \verb{[4]} difficulty in finding a new job
\item \verb{[5]} union members
\item \verb{[6]} older employees
\item \verb{[7]} younger employees
\item \verb{[8]} men
\item \verb{[9]} women
\item \verb{[10]} whites
\item \verb{[11]} minorities-non-whites
\item \verb{[12]} other -please specify
\item \verb{[13]} owners,bosses
\item \verb{[14]} essential employees
\item \verb{[15]} newest employees
\item \verb{[16]} most loyal employees
\item \verb{[17]} production. hands-on employees
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1884/vshow}{GSS Data Explorer page for \code{layoffs}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab those who have worked longest \tab those who have the greatest needs \tab those who are the best workers \tab difficulty in finding a new job \tab union members \tab older employees \tab younger employees \tab men \tab minorities-non-whites \tab other -please specify \tab owners,bosses \tab essential employees \tab newest employees \tab most loyal employees \tab production. hands-on employees \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 685 \tab 98 \tab 529 \tab 3 \tab 4 \tab 11 \tab 3 \tab 5 \tab 7 \tab 22 \tab 9 \tab 4 \tab 2 \tab 3 \tab 4 \tab 48 \tab 7 \tab 2904 \cr
   Total \tab 1460 \tab 685 \tab 98 \tab 529 \tab 3 \tab 4 \tab 11 \tab 3 \tab 5 \tab 7 \tab 22 \tab 9 \tab 4 \tab 2 \tab 3 \tab 4 \tab 48 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decbed}},
\code{\link{decgift}}
}
\concept{Markets}
\keyword{variable}
