% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{ldcgrn}
\alias{ldcgrn}
\title{Poor countries less than rich for environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
ldcgrn
}
\details{
Question How much do you agree or disagree with each of these statements?
B. Poorer countries should be expected to make less effort than richer countries to protect the environment.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4788/vshow}{GSS Data Explorer page for \code{ldcgrn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 40 \tab 164 \tab 245 \tab 522 \tab 142 \tab 60 \tab 103 \tab 2817 \cr
   2010 \tab 614 \tab 31 \tab 202 \tab 202 \tab 737 \tab 173 \tab 78 \tab 7 \tab 2044 \cr
   Total \tab 2155 \tab 71 \tab 366 \tab 447 \tab 1259 \tab 315 \tab 138 \tab 110 \tab 4861 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{econgrn}},
\code{\link{excldimm}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{nukeacc}}
}
\concept{ISSP Environment}
\keyword{variable}
