% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{ldctax}
\alias{ldctax}
\title{Rich countries pay tax to help poor countries}
\source{
General Social Survey https://gss.norc.org
}
\description{
ldctax
}
\details{
Question Turning to international differences, do you agree or disagree:
B. People in wealthy countries should make an additional tax contribution to help people in poor countries.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4325/vshow}{GSS Data Explorer page for \code{ldctax}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 139 \tab 287 \tab 307 \tab 295 \tab 140 \tab 55 \tab 49 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 150 \tab 335 \tab 548 \tab 438 \tab 251 \tab 109 \tab 154 \tab 18 \tab 4032 \cr
   Total \tab 3574 \tab 289 \tab 622 \tab 855 \tab 733 \tab 391 \tab 164 \tab 203 \tab 18 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
