% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{learnhme}
\alias{learnhme}
\title{How learned the availability of the home}
\source{
General Social Survey https://gss.norc.org
}
\description{
learnhme
}
\details{
Question How did you learn about the availability of the home you purchased?
}
\section{Values}{

\itemize{
\item \code{1} from a relative, including in-laws
\item \code{2} a friend or acquaintance
\item \code{3} a friend of a friend or relative, etc
\item \code{4} from an advertisement or sign
\item \code{5} from a real estate agent
\item \code{6} i built it myself
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1814/vshow}{GSS Data Explorer page for \code{learnhme}}.
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
