% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{learning}
\alias{learning}
\title{Importance of learning from co-workers for you job}
\source{
General Social Survey https://gss.norc.org
}
\description{
learning
}
\details{
Question How important were each of the following as sources of the skills that you use in your job?  Would you say very important, important, somewhat important, or not at all important?
C. Informal learning from co-workers or supervisors where you work now.
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} somewhat important
\item \code{4} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1370/vshow}{GSS Data Explorer page for \code{learning}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 295 \tab 353 \tab 145 \tab 101 \tab 6 \tab 12 \tab 1517 \cr
   Total \tab 605 \tab 295 \tab 353 \tab 145 \tab 101 \tab 6 \tab 12 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{baseofop}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{splocnum}},
\code{\link{training}},
\code{\link{wherewrk}}
}
\concept{Work Organizations}
\keyword{variable}
