% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{learnjob}
\alias{learnjob}
\title{How did r find out about present work}
\source{
General Social Survey https://gss.norc.org
}
\description{
learnjob
}
\details{
Question There are many ways people hear about jobs - from other people, from advertisements or employment agencies, and so on. Please indicate how you first found out about work at your present employer. IF YOU ARE NOT CURRENTLY WORKING FOR PAY, PLEASE ANSWER THIS QUESTION FOR YOUR LAST JOB.
}
\section{Values}{

\itemize{
\item \code{1} i have never worked for pay
\item \code{2} from parents, brothers or sisters
\item \code{3} from other relatives
\item \code{4} from a close friend
\item \code{5} from an acquaintance
\item \code{6} from a public employment agency or service
\item \code{7} from a private employment agency
\item \code{8} from a school or university placement office
\item \code{9} from an advertisement or a sign
\item \code{10} the employer contacted me about a job
\item \code{11} i just called them or went there to ask for work
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4224/vshow}{GSS Data Explorer page for \code{learnjob}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab i have never worked for pay \tab from parents, brothers or sisters \tab from other relatives \tab from a close friend \tab from an acquaintance \tab from a public employment agency or service \tab from a private employment agency \tab from a school or university placement office \tab from an advertisement or a sign \tab the employer contacted me about a job \tab i just called them or went there to ask for work \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 18 \tab 58 \tab 56 \tab 218 \tab 145 \tab 62 \tab 24 \tab 55 \tab 177 \tab 103 \tab 179 \tab 29 \tab 25 \tab 2765 \cr
   Total \tab 1616 \tab 18 \tab 58 \tab 56 \tab 218 \tab 145 \tab 62 \tab 24 \tab 55 \tab 177 \tab 103 \tab 179 \tab 29 \tab 25 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{helpjob}},
\code{\link{kidpars}},
\code{\link{lentto}},
\code{\link{talkedto}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
