% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{learnnew}
\alias{learnnew}
\title{Job requires r to learn new things}
\source{
General Social Survey https://gss.norc.org
}
\description{
learnnew
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
A. My job requires that I keep learning new things
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2774/vshow}{GSS Data Explorer page for \code{learnnew}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 748 \tab 782 \tab 177 \tab 70 \tab 6 \tab 13 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 746 \tab 691 \tab 217 \tab 62 \tab 8 \tab 10 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 530 \tab 482 \tab 120 \tab 32 \tab - \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 530 \tab 551 \tab 142 \tab 18 \tab 3 \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 599 \tab 643 \tab 139 \tab 34 \tab 2 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 852 \tab 849 \tab 203 \tab 56 \tab 8 \tab 13 \tab 17 \tab 3544 \cr
   Total \tab 8336 \tab 4005 \tab 3998 \tab 998 \tab 272 \tab 27 \tab 96 \tab 17 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{hrsrelax}},
\code{\link{secondwk}},
\code{\link{whywkhme}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}}
}
\concept{Quality of Working Life}
\keyword{variable}
