% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_261.R
\name{leastcom}
\alias{leastcom}
\title{Have least in common with which group}
\source{
General Social Survey https://gss.norc.org
}
\description{
leastcom
}
\details{
Question Of these groups–whites, blacks, Jews, Hispanics and Latin Americans, Asian Americans, if you had to say, which One (other than your own) do you feel you have the least in common with?
}
\section{Values}{

\itemize{
\item \code{1} whites
\item \code{2} blacks/african-americans
\item \code{3} jews
\item \code{4} hispanics
\item \code{5} asians
\item \code{6} equal in common to all
\item \code{7} nothing in common with any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2361/vshow}{GSS Data Explorer page for \code{leastcom}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab whites \tab blacks/african-americans \tab jews \tab hispanics \tab asians \tab equal in common to all \tab nothing in common with any \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 35 \tab 210 \tab 174 \tab 173 \tab 426 \tab 164 \tab 71 \tab 107 \tab 38 \tab 2817 \cr
   Total \tab 1419 \tab 35 \tab 210 \tab 174 \tab 173 \tab 426 \tab 164 \tab 71 \tab 107 \tab 38 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{neieth1}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
