% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{leavejb}
\alias{leavejb}
\title{Why did r leave last job}
\source{
General Social Survey https://gss.norc.org
}
\description{
leavejb
}
\details{
Question Why did you leave your last (full-time) job? Did you leave voluntarily, was the job abolished, or were you permanently laid off?
}
\section{Values}{

\itemize{
\item \code{1} left voluntarily
\item \code{2} job was abolished
\item \code{3} permanently laid off
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3894/vshow}{GSS Data Explorer page for \code{leavejb}}.
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}

Other Employment: 
\code{\link{autonojb}},
\code{\link{lastyrwk}},
\code{\link{whynojb1}},
\code{\link{whynojb2}}

Other Job: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}

Other Unemployment: 
\code{\link{autonojb}},
\code{\link{lastyrwk}},
\code{\link{whynojb1}},
\code{\link{whynojb2}}
}
\concept{Employment}
\concept{Global Economics}
\concept{Job}
\concept{Unemployment}
\keyword{variable}
