% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{leftrght}
\alias{leftrght}
\title{How left or right in politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
leftrght
}
\details{
Question In politics people sometimes talk of left and right. Where would you place yourself on a scale from 0 to 10 where 0 means the left and 10 means the right?
}
\section{Values}{

\itemize{
\item \verb{[0]} left
\item \verb{[10]} right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4971/vshow}{GSS Data Explorer page for \code{leftrght}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab left \tab 1 \tab right \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2014 \tab 1274 \tab 43 \tab 21 \tab 70 \tab 51 \tab 67 \tab 87 \tab 447 \tab 92 \tab 85 \tab 81 \tab 21 \tab 162 \tab 37 \tab - \tab 2538 \cr
   2024 \tab 27 \tab 146 \tab 72 \tab 268 \tab 164 \tab 205 \tab 205 \tab 851 \tab 207 \tab 241 \tab 184 \tab 108 \tab 559 \tab 72 \tab - \tab 3309 \cr
   Total \tab 1301 \tab 189 \tab 93 \tab 338 \tab 215 \tab 272 \tab 292 \tab 1298 \tab 299 \tab 326 \tab 265 \tab 129 \tab 721 \tab 109 \tab 0 \tab 5847 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{hlthall}},
\code{\link{notvote}},
\code{\link{ntcitvte}}

Other Political: 
\code{\link{ntcitvte}}
}
\concept{ISSP Citizenship}
\concept{Political}
\keyword{variable}
