% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{lessfare}
\alias{lessfare}
\title{If less welfare more will work}
\source{
General Social Survey https://gss.norc.org
}
\description{
lessfare
}
\details{
Question Here are some proposals about welfare. I would like to know if you favor or oppose them. (READ "A") - Would you say that you strongly favor it, favor it, neither favor nor oppose it, oppose it, or strongly oppose it?
B. Second, reducing welfare benefits to make working for a living more attractive.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1229/vshow}{GSS Data Explorer page for \code{lessfare}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 445 \tab 494 \tab 149 \tab 184 \tab 53 \tab 38 \tab 9 \tab 1372 \cr
   Total \tab 0 \tab 445 \tab 494 \tab 149 \tab 184 \tab 53 \tab 38 \tab 9 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{workfare}}
}
\concept{Intergroup Relations}
\keyword{variable}
