% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{lesspain}
\alias{lesspain}
\title{Feel deep sense of resp for reducing pain}
\source{
General Social Survey https://gss.norc.org
}
\description{
lesspain
}
\details{
Question I feel a deep sense of responsibility for reducing pain and suffering in the world.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2103/vshow}{GSS Data Explorer page for \code{lesspain}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 239 \tab 640 \tab 433 \tab 97 \tab 12 \tab 24 \tab 2832 \cr
   Total \tab 1387 \tab 239 \tab 640 \tab 433 \tab 97 \tab 12 \tab 24 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{relcmfrt}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
