% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{lessprd}
\alias{lessprd}
\title{Agree often less proud of america}
\source{
General Social Survey https://gss.norc.org
}
\description{
lessprd
}
\details{
Question I am often less proud of America than I would like to be.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4882/vshow}{GSS Data Explorer page for \code{lessprd}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 71 \tab 314 \tab 256 \tab 453 \tab 107 \tab 14 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 74 \tab 400 \tab 259 \tab 428 \tab 72 \tab 35 \tab 6 \tab 2538 \cr
   Total \tab 2860 \tab 145 \tab 714 \tab 515 \tab 881 \tab 179 \tab 49 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
