% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{letdie1y}
\alias{letdie1y}
\title{Allow incurable patients to die (form 2)}
\source{
General Social Survey https://gss.norc.org
}
\description{
letdie1y
}
\details{
Question When a person has a disease that cannot be cured, do you think doctors should or should not be allowed by law to end the patient’s life by some painless means if the patient and his family request it?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5573/vshow}{GSS Data Explorer page for \code{letdie1y}}.
}

\seealso{
Other Form Y: 
\code{\link{polhitoy}},
\code{\link{popespky}}
}
\concept{Euthanasia}
\concept{Form Y}
\keyword{variable}
