% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{letin}
\alias{letin}
\title{Increase or decrease immigrants to us}
\source{
General Social Survey https://gss.norc.org
}
\description{
letin
}
\details{
Question Do you think the number of immigrants from foreign countries who are permitted to come to the United States to live should be increased a lot, increased a little, left the same as it is now, decreased a little, or decreased a lot?
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} left the same as it is now
\item \code{4} decreased a little
\item \code{5} decreased a lot
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1611/vshow}{GSS Data Explorer page for \code{letin}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab left the same as it is now \tab decreased a little \tab decreased a lot \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 35 \tab 53 \tab 395 \tab 407 \tab 507 \tab 69 \tab 8 \tab 2992 \cr
   2000 \tab 1419 \tab 50 \tab 74 \tab 612 \tab 281 \tab 301 \tab 76 \tab 4 \tab 2817 \cr
   Total \tab 2937 \tab 85 \tab 127 \tab 1007 \tab 688 \tab 808 \tab 145 \tab 12 \tab 5809 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
