% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{letin1a}
\alias{letin1a}
\title{Number of immigrants nowadays should be}
\source{
General Social Survey https://gss.norc.org
}
\description{
letin1a
}
\details{
Question Do you think the number of immigrants to America nowadays should be…
}
\section{Values}{

\itemize{
\item \verb{[1]} increased a lot
\item \verb{[2]} increased a little
\item \verb{[3]} remain the same as it is
\item \verb{[4]} reduced a little
\item \verb{[5]} reduced a lot
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4901/vshow}{GSS Data Explorer page for \code{letin1a}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab remain the same as it is \tab reduced a little \tab reduced a lot \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2004 \tab 1906 \tab 30 \tab 52 \tab 330 \tab 222 \tab 256 \tab 16 \tab - \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 76 \tab 170 \tab 677 \tab 464 \tab 558 \tab 43 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 47 \tab 105 \tab 448 \tab 307 \tab 387 \tab 30 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 62 \tab 138 \tab 489 \tab 345 \tab 359 \tab 29 \tab 8 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 58 \tab 117 \tab 505 \tab 288 \tab 294 \tab 36 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 76 \tab 166 \tab 654 \tab 375 \tab 353 \tab 46 \tab 5 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 108 \tab 218 \tab 742 \tab 419 \tab 358 \tab 32 \tab 11 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 113 \tab 227 \tab 646 \tab 298 \tab 236 \tab 32 \tab 7 \tab - \tab 2348 \cr
   2021 \tab 1320 \tab 227 \tab 476 \tab 1128 \tab 426 \tab 413 \tab 8 \tab 3 \tab 31 \tab 4032 \cr
   2022 \tab - \tab 398 \tab 542 \tab 1288 \tab 593 \tab 619 \tab 73 \tab 17 \tab 14 \tab 3544 \cr
   Total \tab 10355 \tab 1195 \tab 2211 \tab 6907 \tab 3737 \tab 3833 \tab 345 \tab 64 \tab 45 \tab 28692 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amproud1}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{belikeus}},
\code{\link{biasnews}},
\code{\link{bizmoney}},
\code{\link{brneffrt}},
\code{\link{citizen}},
\code{\link{citworld}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{decsorgs}},
\code{\link{ethclose}},
\code{\link{experts}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{freetrde}},
\code{\link{fulldem}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immassim}},
\code{\link{immcosts}},
\code{\link{immcrime1}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{immideas}},
\code{\link{immideas1}},
\code{\link{immimp}},
\code{\link{immjobs1}},
\code{\link{immrghts}},
\code{\link{immwlfare}},
\code{\link{imports}},
\code{\link{incgap1}},
\code{\link{internet}},
\code{\link{intlincs}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{lessprd}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{refugees}},
\code{\link{res161}},
\code{\link{shortcom}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{wrldgovt}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{ISSP National Identity}
\concept{Immigration}
\keyword{variable}
