% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{letinasn1}
\alias{letinasn1}
\title{Immigration from asia should decrease or increase}
\source{
General Social Survey https://gss.norc.org
}
\description{
letinasn1
}
\details{
Question Do you think the number of immigrants from Asia who are permitted to come to the United States to live should be decreased, increased or left the same?
}
\section{Values}{

\itemize{
\item \code{1} decreased
\item \code{2} increased
\item \code{3} left the same
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8004/vshow}{GSS Data Explorer page for \code{letinasn1}}.
}

\keyword{variable}
