% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{letinhsp}
\alias{letinhsp}
\title{Immigration from latin america should be increased}
\source{
General Social Survey https://gss.norc.org
}
\description{
letinhsp
}
\details{
Question What about the number of immigrants from Latin America (that is, Spanish-speaking countries of the Americas)–should it be increased a lot, increased a little, left the same as it is now, decreased a little, or decreased a lot?
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} left the same as it is now
\item \code{4} decreased a little
\item \code{5} decreased a lot
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2304/vshow}{GSS Data Explorer page for \code{letinhsp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab left the same as it is now \tab decreased a little \tab decreased a lot \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 41 \tab 77 \tab 615 \tab 279 \tab 303 \tab 78 \tab 5 \tab 2817 \cr
   Total \tab 1419 \tab 41 \tab 77 \tab 615 \tab 279 \tab 303 \tab 78 \tab 5 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{othhome}},
\code{\link{spklangw}},
\code{\link{uswht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
