% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{letinhsp1}
\alias{letinhsp1}
\title{Immigration from latin america should decrease or increase}
\source{
General Social Survey https://gss.norc.org
}
\description{
letinhsp1
}
\details{
Question Do you think the number of immigrants from Latin America who are permitted to come to the United States to live should be decreased, increased or left the same?
}
\section{Values}{

\itemize{
\item \code{1} decreased
\item \code{2} increased
\item \code{3} left the same
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8005/vshow}{GSS Data Explorer page for \code{letinhsp1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab decreased \tab increased \tab left the same \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1203 \tab 513 \tab 473 \tab 1246 \tab 84 \tab 7 \tab 18 \tab 3544 \cr
   Total \tab 1203 \tab 513 \tab 473 \tab 1246 \tab 84 \tab 7 \tab 18 \tab 3544 \cr
}
}

\keyword{variable}
