% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{levels}
\alias{levels}
\title{How many levels of supervision beneath you}
\source{
General Social Survey https://gss.norc.org
}
\description{
levels
}
\details{
Question In total, how many levels of supervision are there beneath yours (including the one immediately under you?)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1316/vshow}{GSS Data Explorer page for \code{levels}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab no answer \tab Total \cr
   1991 \tab 1411 \tab 36 \tab 34 \tab 15 \tab 5 \tab 2 \tab 1 \tab 4 \tab 9 \tab 1517 \cr
   Total \tab 1411 \tab 36 \tab 34 \tab 15 \tab 5 \tab 2 \tab 1 \tab 4 \tab 9 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{totsup}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
