% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{lfechnce}
\alias{lfechnce}
\title{Life is a matter of chance}
\source{
General Social Survey https://gss.norc.org
}
\description{
lfechnce
}
\details{
Question I'm going to read some statements that give reasons why a person's life turns out well or poorly. (HAND CARD BZ TO RESPONDENT.)  As I read each one, tell me whether you think it is very important, important, somewhat important, or not at all important for how somebody's life turns out? E. It's just a matter of chance.
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} somewhat important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1386/vshow}{GSS Data Explorer page for \code{lfechnce}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1993 \tab - \tab 55 \tab 250 \tab 591 \tab 668 \tab 32 \tab 10 \tab - \tab 1606 \cr
   Total \tab 0 \tab 55 \tab 250 \tab 591 \tab 668 \tab 32 \tab 10 \tab 0 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{bigband}},
\code{\link{blues}},
\code{\link{blugrass}},
\code{\link{camping}},
\code{\link{classicl}},
\code{\link{classics}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{conrock}},
\code{\link{country}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{english}},
\code{\link{excelart}},
\code{\link{folk}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{gospel}},
\code{\link{grtbooks}},
\code{\link{hosthome}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{huntfish}},
\code{\link{hvymetal}},
\code{\link{impcultr}},
\code{\link{impfinan}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{jazz}},
\code{\link{judgeart}},
\code{\link{latin}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{makeart}},
\code{\link{modpaint}},
\code{\link{moodeasy}},
\code{\link{mostimp1}},
\code{\link{mostimp2}},
\code{\link{mostimp3}},
\code{\link{mostimp4}},
\code{\link{musicals}},
\code{\link{newage}},
\code{\link{obtohelp}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{opoutcme}},
\code{\link{pclit}},
\code{\link{perform}},
\code{\link{plymusic}},
\code{\link{popmusic}},
\code{\link{rap}},
\code{\link{readfict}},
\code{\link{reggae}},
\code{\link{richpoor}},
\code{\link{seemovie}},
\code{\link{selfirst}},
\code{\link{standup}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}},
\code{\link{united}},
\code{\link{usevcr}},
\code{\link{visitart}}

Other Success: 
\code{\link{balwkfam}},
\code{\link{famlife}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{worklife}}
}
\concept{Culture}
\concept{Success}
\keyword{variable}
