% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{lfesocty}
\alias{lfesocty}
\title{Society gives some head start & holds back others}
\source{
General Social Survey https://gss.norc.org
}
\description{
lfesocty
}
\details{
Question I'm going to read some statements that give reasons why a person's life turns out well or poorly. (HAND CARD BZ TO RESPONDENT.)  As I read each one, tell me whether you think it is very important, important, somewhat important, or not at all important for how somebody's life turns out?
C. Society gives some people a head start and holds others back.
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} somewhat important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1384/vshow}{GSS Data Explorer page for \code{lfesocty}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 213 \tab 598 \tab 524 \tab 215 \tab 44 \tab 12 \tab 1606 \cr
   Total \tab 0 \tab 213 \tab 598 \tab 524 \tab 215 \tab 44 \tab 12 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{bigband}},
\code{\link{blugrass}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{obtohelp}},
\code{\link{opoutcme}},
\code{\link{united}}
}
\concept{Culture}
\keyword{variable}
