% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{lftplane}
\alias{lftplane}
\title{Key factor that enables airplane to lift}
\source{
General Social Survey https://gss.norc.org
}
\description{
lftplane
}
\details{
Question Which of the following is a key factor that enables an airplane to lift?
}
\section{Values}{

\itemize{
\item \verb{[1]} air pressure beneath the wing is greater than that above the wing
\item \verb{[2]} pressure within the airplane is greater than that of the outside
\item \verb{[3]} engine power is greater than that of friction
\item \verb{[4]} the plane's wing is lighter than air
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3473/vshow}{GSS Data Explorer page for \code{lftplane}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab air pressure beneath the wing is greater than that above the wing \tab pressure within the airplane is greater than that of the outside \tab engine power is greater than that of friction \tab the plane's wing is lighter than air \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 799 \tab 90 \tab 345 \tab 40 \tab 211 \tab 20 \tab 2023 \cr
   Total \tab 518 \tab 799 \tab 90 \tab 345 \tab 40 \tab 211 \tab 20 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
