% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{libathy}
\alias{libathy}
\title{Allow anti-religionist to teach y}
\source{
General Social Survey https://gss.norc.org
}
\description{
libathy
}
\details{
Question (There are always some people whose ideas are considered bad or dangerous by other people. For instance, somebody who is against all churches and religion…)
If some people in your community suggested that a book this person wrote against churches and religion should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} remove
\item \verb{[2]} not remove
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7435/vshow}{GSS Data Explorer page for \code{libathy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab remove \tab not remove \tab skipped on web \tab don't know \tab no answer \tab Total \cr
   2021 \tab 2675 \tab 199 \tab 1150 \tab 8 \tab - \tab - \tab 4032 \cr
   2022 \tab 2367 \tab 208 \tab 945 \tab 2 \tab 19 \tab 3 \tab 3544 \cr
   Total \tab 5042 \tab 407 \tab 2095 \tab 10 \tab 19 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
