% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{libcom}
\alias{libcom}
\title{Allow communists book in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libcom
}
\details{
Question Now, I should like to ask you some questions about a man who admits he is a Communist.
C. Suppose he wrote a book which is in your public library. Somebody in your community suggests that the book should be removed from the library. Would you favor removing it, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} remove
\item \verb{[2]} not remove
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/257/vshow}{GSS Data Explorer page for \code{libcom}}.

Counts by year:\tabular{llllllll}{
   year \tab remove \tab not remove \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 681 \tab 851 \tab 76 \tab 5 \tab - \tab - \tab 1613 \cr
   1973 \tab 587 \tab 876 \tab 35 \tab 6 \tab - \tab - \tab 1504 \cr
   1974 \tab 561 \tab 866 \tab 51 \tab 6 \tab - \tab - \tab 1484 \cr
   1976 \tab 601 \tab 838 \tab 55 \tab 5 \tab - \tab - \tab 1499 \cr
   1977 \tab 642 \tab 842 \tab 39 \tab 7 \tab - \tab - \tab 1530 \cr
   1980 \tab 571 \tab 839 \tab 55 \tab 3 \tab - \tab - \tab 1468 \cr
   1982 \tab 738 \tab 1001 \tab 107 \tab 14 \tab - \tab - \tab 1860 \cr
   1984 \tab 545 \tab 873 \tab 49 \tab 6 \tab - \tab - \tab 1473 \cr
   1985 \tab 597 \tab 874 \tab 57 \tab 6 \tab - \tab - \tab 1534 \cr
   1987 \tab 671 \tab 1078 \tab 60 \tab 10 \tab - \tab - \tab 1819 \cr
   1988 \tab 365 \tab 579 \tab 30 \tab 3 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 350 \tab 633 \tab 44 \tab 6 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 299 \tab 584 \tab 31 \tab 3 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 293 \tab 667 \tab 29 \tab 4 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 310 \tab 713 \tab 46 \tab 6 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 616 \tab 1322 \tab 54 \tab 4 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 598 \tab 1241 \tab 79 \tab 5 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 538 \tab 1247 \tab 90 \tab 7 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 569 \tab 1223 \tab 62 \tab 7 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 263 \tab 636 \tab 22 \tab 3 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 251 \tab 630 \tab 14 \tab 3 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 597 \tab 1364 \tab 40 \tab 2 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 409 \tab 900 \tab 41 \tab 2 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 367 \tab 876 \tab 34 \tab 4 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 335 \tab 919 \tab 51 \tab 3 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 448 \tab 1220 \tab 39 \tab 7 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 480 \tab 1350 \tab 47 \tab 13 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 395 \tab 1110 \tab 62 \tab 7 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 278 \tab 1025 \tab 2 \tab - \tab 2714 \tab 13 \tab 4032 \cr
   2022 \tab 307 \tab 810 \tab 35 \tab 4 \tab 2380 \tab 8 \tab 3544 \cr
   Total \tab 14262 \tab 27987 \tab 1436 \tab 161 \tab 22432 \tab 21 \tab 66299 \cr
}
}

\seealso{
Other Core: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Core}
\keyword{variable}
