% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_624.R
\name{libhomoy}
\alias{libhomoy}
\title{Allow homosexuals book in library y}
\source{
General Social Survey https://gss.norc.org
}
\description{
libhomoy
}
\details{
Question If somebody in your community suggests that a book the gay person wrote in favor of homosexuality should be taken out of your public library, would you favor  removing it, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} remove
\item \verb{[2]} not remove
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7444/vshow}{GSS Data Explorer page for \code{libhomoy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab remove \tab not remove \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2675 \tab 149 \tab 1193 \tab 4 \tab 11 \tab 4032 \cr
   Total \tab 2675 \tab 149 \tab 1193 \tab 4 \tab 11 \tab 4032 \cr
}
}

\keyword{variable}
