% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{libmil}
\alias{libmil}
\title{Allow militarists book in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libmil
}
\details{
Question Consider a person who advocates doing away with elections and letting the military run the country.
C. Suppose he wrote a book advocating doing away with elections and letting the military run the country. somebody in your community suggests that the book be removed from the public library. Would you favor removing it, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
   1998 \tab A/C \tab partial \cr
   2000 \tab A/C \tab partial \cr
   2002 \tab A/C \tab partial \cr
   2004 \tab A/C \tab partial \cr
   2006 \tab A/C \tab partial \cr
   2008 \tab A/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
   2012 \tab A/C \tab partial \cr
   2014 \tab A/C \tab partial \cr
   2016 \tab A/C \tab partial \cr
   2018 \tab A/C \tab partial \cr
   2021 \tab A/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/260/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab don't know \tab no answer \tab not remove \tab remove \tab skipped on web \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 52 \tab 6 \tab 838 \tab 603 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 41 \tab 4 \tab 833 \tab 652 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 49 \tab 2 \tab 852 \tab 565 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 88 \tab 6 \tab 977 \tab 789 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 30 \tab 5 \tab 869 \tab 569 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 33 \tab 5 \tab 862 \tab 634 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 65 \tab 8 \tab 1020 \tab 726 \tab - \tab - \tab 1819 \cr
   1988 \tab 504 \tab 36 \tab 7 \tab 551 \tab 383 \tab - \tab - \tab 1481 \cr
   1989 \tab 504 \tab 39 \tab 7 \tab 610 \tab 377 \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 34 \tab 4 \tab 554 \tab 325 \tab - \tab - \tab 1372 \cr
   1991 \tab 524 \tab 22 \tab 5 \tab 661 \tab 305 \tab - \tab - \tab 1517 \cr
   1993 \tab 531 \tab 37 \tab 4 \tab 721 \tab 313 \tab - \tab - \tab 1606 \cr
   1994 \tab 996 \tab 57 \tab 3 \tab 1274 \tab 662 \tab - \tab - \tab 2992 \cr
   1996 \tab 981 \tab 68 \tab 4 \tab 1234 \tab 617 \tab - \tab - \tab 2904 \cr
   1998 \tab 950 \tab 90 \tab 3 \tab 1253 \tab 536 \tab - \tab - \tab 2832 \cr
   2000 \tab 956 \tab 65 \tab 6 \tab 1187 \tab 603 \tab - \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 24 \tab 2 \tab 642 \tab 256 \tab - \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 15 \tab 3 \tab 615 \tab 265 \tab - \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 35 \tab 2 \tab 1375 \tab 591 \tab - \tab - \tab 4510 \cr
   2008 \tab 671 \tab 40 \tab 4 \tab 916 \tab 392 \tab - \tab - \tab 2023 \cr
   2010 \tab 763 \tab 21 \tab 2 \tab 901 \tab 357 \tab - \tab - \tab 2044 \cr
   2012 \tab 666 \tab 31 \tab 6 \tab 927 \tab 344 \tab - \tab - \tab 1974 \cr
   2014 \tab 824 \tab 28 \tab 9 \tab 1197 \tab 480 \tab - \tab - \tab 2538 \cr
   2016 \tab 977 \tab 31 \tab 8 \tab 1377 \tab 474 \tab - \tab - \tab 2867 \cr
   2018 \tab 774 \tab 33 \tab 4 \tab 1135 \tab 402 \tab - \tab - \tab 2348 \cr
   2021 \tab 2714 \tab 1 \tab 1 \tab 910 \tab 392 \tab 14 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 30744 \tab 1065 \tab 120 \tab 24291 \tab 12612 \tab 14 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}

Other Civil Liberties: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Civil Liberties}
\concept{Core}
\keyword{variable}
