% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{libmslm}
\alias{libmslm}
\title{Allow anti-american muslim clergymen's books in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libmslm
}
\details{
Question Now consider a Muslim clergyman who preaches hatred of the United States.
C. If some people in your community suggested that a book he wrote which preaches hatred of the United States should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/266/vshow}{GSS Data Explorer page for \code{libmslm}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab remove \tab not remove \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 671 \tab 663 \tab 662 \tab 21 \tab 6 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 649 \tab 619 \tab 9 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 641 \tab 638 \tab 22 \tab 7 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 858 \tab 831 \tab 14 \tab 11 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 935 \tab 930 \tab 16 \tab 9 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 775 \tab 776 \tab 17 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 2714 \tab 569 \tab 735 \tab 3 \tab - \tab 11 \tab 4032 \cr
   2022 \tab 2380 \tab 513 \tab 626 \tab 18 \tab 1 \tab 6 \tab 3544 \cr
   Total \tab 9769 \tab 5603 \tab 5817 \tab 120 \tab 44 \tab 17 \tab 21370 \cr
}
}

\seealso{
Other Core: 
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libhomo}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}
}
\concept{Core}
\keyword{variable}
