% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_624.R
\name{libmslmy}
\alias{libmslmy}
\title{Allow anti-american muslim clergymen's books in library y}
\source{
General Social Survey https://gss.norc.org
}
\description{
libmslmy
}
\details{
Question (Now consider a Islamic religious leader who preaches hatred of the United States…)
If some people in your community suggested that a book this person wrote which preaches hatred of the United States should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} remove
\item \verb{[2]} not remove
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7446/vshow}{GSS Data Explorer page for \code{libmslmy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab remove \tab not remove \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 2675 \tab 616 \tab 726 \tab 4 \tab 11 \tab - \tab 4032 \cr
   2022 \tab 2367 \tab 530 \tab 623 \tab 16 \tab 6 \tab 2 \tab 3544 \cr
   Total \tab 5042 \tab 1146 \tab 1349 \tab 20 \tab 17 \tab 2 \tab 7576 \cr
}
}

\keyword{variable}
