% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{libracy}
\alias{libracy}
\title{Allow racists book in library y}
\source{
General Social Survey https://gss.norc.org
}
\description{
libracy
}
\details{
Question (Or consider a person who believes that Black people are genetically inferior…)If some people in your community suggested that a book this person wrote which said Black people are inferior should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7437/vshow}{GSS Data Explorer page for \code{libracy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab remove \tab not remove \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 2675 \tab 539 \tab 804 \tab 2 \tab 12 \tab - \tab 4032 \cr
   2022 \tab 2367 \tab 480 \tab 669 \tab 20 \tab 5 \tab 3 \tab 3544 \cr
   Total \tab 5042 \tab 1019 \tab 1473 \tab 22 \tab 17 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
