% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{libsoc}
\alias{libsoc}
\title{Allow socialists book in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libsoc
}
\details{
Question Or consider a person who favored government ownership of all the railroads and all big industries.
C. If some people in your community suggested a book he wrote favoring government ownership should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973, 1974 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/251/vshow}{GSS Data Explorer page for \code{libsoc}}.

Counts by year:\tabular{lllllll}{
   year \tab remove \tab not remove \tab don't know \tab no answer \tab iap \tab Total \cr
   1972 \tab 409 \tab 1086 \tab 107 \tab 11 \tab - \tab 1613 \cr
   1973 \tab 372 \tab 1070 \tab 59 \tab 3 \tab - \tab 1504 \cr
   1974 \tab 380 \tab 1025 \tab 74 \tab 5 \tab - \tab 1484 \cr
   Total \tab 1161 \tab 3181 \tab 240 \tab 19 \tab 0 \tab 4601 \cr
}
}

\seealso{
Other Core: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Core}
\keyword{variable}
