% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_572.R
\name{lietest}
\alias{lietest}
\title{People with secret should be subject to lie tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
lietest
}
\details{
Question Protecting secrets is a continuing concern of the government.
I am going to ask you about various security measures that the government might apply to individuals with a SECRET of TOP SECRET clearance. Do you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree that people with SECRET or TOP SECRET clearance should be subject to...
A. Periodic lie detector tests
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5243/vshow}{GSS Data Explorer page for \code{lietest}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 398 \tab 632 \tab 137 \tab 137 \tab 26 \tab 41 \tab 16 \tab 2832 \cr
   2000 \tab 1398 \tab 458 \tab 628 \tab 151 \tab 111 \tab 29 \tab 26 \tab 16 \tab 2817 \cr
   Total \tab 2843 \tab 856 \tab 1260 \tab 288 \tab 248 \tab 55 \tab 67 \tab 32 \tab 5649 \cr
}
}

\seealso{
Other National Security: 
\code{\link{bugging}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{compfin}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{testdrug}}
}
\concept{National Security}
\keyword{variable}
