% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_313.R
\name{lifein5}
\alias{lifein5}
\title{R's rating of life overall in five years from 0-10}
\source{
General Social Survey https://gss.norc.org
}
\description{
lifein5
}
\details{
Question If you were asked to rate your overall life from 0-10, where 10 represents the best possible state and 0 represents the worst possible state, how do you think you would rate your life five years from today?
}
\section{Values}{

\itemize{
\item \code{0} 0 - worst possible state
\item \code{10} 10 - best possible state
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7157/vshow}{GSS Data Explorer page for \code{lifein5}}.
}

\keyword{variable}
