% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{likediff}
\alias{likediff}
\title{Afterlife contrast: better or lacking}
\source{
General Social Survey https://gss.norc.org
}
\description{
likediff
}
\details{
Question A. People picture life after death in many different ways. We'd like to know how you think of life after death. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of life after death. Look at the first set of contrasting images. If you think that life after death is "A life like the one here on earth only better," you would place yourself at 1. If you think it is "A life without many things which make our present life enjoyable" you would place yourself at 7. If you think it is somewhere between these two you would place yourself at 2, 3, 4, 5, or 6. Where would you place your images of life after death for . . .?
}
\section{Values}{

\itemize{
\item \code{1} better than earth
\item \code{7} life lacking
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/311/vshow}{GSS Data Explorer page for \code{likediff}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab better than earth \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab life lacking \tab don't know \tab no answer \tab Total \cr
   1986 \tab 246 \tab 503 \tab 125 \tab 111 \tab 224 \tab 68 \tab 32 \tab 114 \tab 21 \tab 26 \tab 1470 \cr
   1987 \tab 369 \tab 584 \tab 148 \tab 132 \tab 226 \tab 77 \tab 64 \tab 132 \tab 26 \tab 61 \tab 1819 \cr
   1988 \tab 282 \tab 492 \tab 116 \tab 100 \tab 224 \tab 69 \tab 35 \tab 100 \tab 16 \tab 47 \tab 1481 \cr
   1989 \tab 749 \tab 310 \tab 86 \tab 77 \tab 137 \tab 45 \tab 27 \tab 60 \tab 10 \tab 36 \tab 1537 \cr
   Total \tab 1646 \tab 1889 \tab 475 \tab 420 \tab 811 \tab 259 \tab 158 \tab 406 \tab 73 \tab 170 \tab 6307 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{grace}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
