% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{likeoth}
\alias{likeoth}
\title{How much do you like or dislike the person}
\source{
General Social Survey https://gss.norc.org
}
\description{
likeoth
}
\details{
Question Thinking generally now, how much do you like or dislike the person/the people who made you angry or annoyed? Do you:
}
\section{Values}{

\itemize{
\item \verb{[1]} like him-her-them a great deal
\item \verb{[2]} like him-her-them a little
\item \verb{[3]} feel neutral toward him-her-them
\item \verb{[4]} dislike him or her-them a little
\item \verb{[5]} dislike him or her-them a lot
\item \verb{[6]} if volunteered - love
\item \verb{[7]} if volunteered - dont know
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1786/vshow}{GSS Data Explorer page for \code{likeoth}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab like him-her-them a great deal \tab like him-her-them a little \tab feel neutral toward him-her-them \tab dislike him or her-them a little \tab dislike him or her-them a lot \tab if volunteered - love \tab if volunteered - dont know \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1908 \tab 301 \tab 108 \tab 255 \tab 92 \tab 114 \tab 63 \tab 29 \tab 1 \tab 33 \tab 2904 \cr
   Total \tab 1908 \tab 301 \tab 108 \tab 255 \tab 92 \tab 114 \tab 63 \tab 29 \tab 1 \tab 33 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
