% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_634.R
\name{limitpay}
\alias{limitpay}
\title{The government should take steps to limit the pay of executives at major compani}
\source{
General Social Survey https://gss.norc.org
}
\description{
limitpay
}
\details{
Question The government should take steps to limit the pay of executives at major companies
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7550/vshow}{GSS Data Explorer page for \code{limitpay}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2244 \tab 259 \tab 432 \tab 459 \tab 374 \tab 173 \tab 71 \tab 2 \tab 18 \tab 4032 \cr
   Total \tab 2244 \tab 259 \tab 432 \tab 459 \tab 374 \tab 173 \tab 71 \tab 2 \tab 18 \tab 4032 \cr
}
}

\keyword{variable}
