% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_273.R
\name{litget}
\alias{litget}
\title{Use web to download poem or fiction}
\source{
General Social Survey https://gss.norc.org
}
\description{
litget
}
\details{
Question C. Downloaded a poem or work of fiction from a Web site to your computer?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2481/vshow}{GSS Data Explorer page for \code{litget}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2554 \tab 72 \tab 76 \tab 115 \tab 2817 \cr
   2002 \tab 2555 \tab 49 \tab 115 \tab 46 \tab 2765 \cr
   Total \tab 5109 \tab 121 \tab 191 \tab 161 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{litauth}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{wwwbiz}},
\code{\link{wwwcowrk}},
\code{\link{wwwgfbf}},
\code{\link{wwwlit}},
\code{\link{wwwpersn}},
\code{\link{wwwsp}}
}
\concept{Information Society}
\keyword{variable}
