% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_566.R
\name{live4}
\alias{live4}
\title{Serious problem with your landlord}
\source{
General Social Survey https://gss.norc.org
}
\description{
live4
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
G. What about things concerning where you live?  During the last year, did any of the following happen?
4. Serious trouble with a landlord or building manager.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5188/vshow}{GSS Data Explorer page for \code{live4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 500 \tab 26 \tab 986 \tab 5 \tab 1517 \cr
   2004 \tab 1472 \tab 31 \tab 1302 \tab 7 \tab 2812 \cr
   Total \tab 1972 \tab 57 \tab 2288 \tab 12 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}}
}
\concept{Negative Life Events}
\keyword{variable}
