% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_130.R
\name{liveasns}
\alias{liveasns}
\title{Neighborhood half asian}
\source{
General Social Survey https://gss.norc.org
}
\description{
liveasns
}
\details{
Question Now I'm going to ask you about different types of contact with various groups of people. In each situation would you please tell me whether you would be very much in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very much opposed to it happening?
C. Living in a neighborhood where half of your neighbors were Asian Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1216/vshow}{GSS Data Explorer page for \code{liveasns}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab - \tab 46 \tab 140 \tab 663 \tab 358 \tab 128 \tab 27 \tab 10 \tab - \tab 1372 \cr
   2000 \tab 1419 \tab 150 \tab 232 \tab 680 \tab 170 \tab 74 \tab 78 \tab 14 \tab - \tab 2817 \cr
   Total \tab 1419 \tab 196 \tab 372 \tab 1343 \tab 528 \tab 202 \tab 105 \tab 24 \tab 0 \tab 4189 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Asian Americans: 
\code{\link{acqasian}},
\code{\link{acqfmasn}},
\code{\link{acqnhasn}},
\code{\link{acqvaasn}},
\code{\link{acqwkasn}},
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asngovt}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{comasn}},
\code{\link{fairasns}},
\code{\link{famasns}},
\code{\link{fareasns}},
\code{\link{feelasns}},
\code{\link{hspasn10}},
\code{\link{influasn}},
\code{\link{intlasns}},
\code{\link{knwasn}},
\code{\link{marasian}},
\code{\link{patrasns}},
\code{\link{trtasian}},
\code{\link{usasn}},
\code{\link{violasns}},
\code{\link{wlthasns}},
\code{\link{workasns}}

Other Residence: 
\code{\link{abroad}},
\code{\link{agecmeus}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{comamind}},
\code{\link{comtype}},
\code{\link{comyear}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwell5}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{geomobil}},
\code{\link{hspasn10}},
\code{\link{livecom}},
\code{\link{livecom1}},
\code{\link{livehome}},
\code{\link{livehsps}},
\code{\link{loclived}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{res16}},
\code{\link{res161}},
\code{\link{res2006}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Asian Americans}
\concept{Intergroup Relations}
\concept{Residence}
\keyword{variable}
