% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{livecom}
\alias{livecom}
\title{Length of residence in city, town, comm}
\source{
General Social Survey https://gss.norc.org
}
\description{
livecom
}
\details{
Question How long have you lived in the city, town or community where you live now? PLEASE FILL IN THE NUMBER OF YEARS IN THE SPACE BELOW. IF LESS THAN ONE YEAR, MAKE A CHECK ON THE LINE NEXT TO THAT CHOICE.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4181/vshow}{GSS Data Explorer page for \code{livecom}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 7 \tab 70 \tab 72 \tab 74 \tab 75 \tab 76 \tab 78 \tab 79 \tab 8 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 89 \tab 9 \tab no answer \tab Total \cr
   1986 \tab 42 \tab 134 \tab 31 \tab 46 \tab 28 \tab 28 \tab 22 \tab 16 \tab 35 \tab 21 \tab 19 \tab 26 \tab 21 \tab 75 \tab 44 \tab 26 \tab 27 \tab 17 \tab 21 \tab 31 \tab 16 \tab 25 \tab 18 \tab 20 \tab 52 \tab 33 \tab 17 \tab 11 \tab 31 \tab 15 \tab 26 \tab 17 \tab 10 \tab 13 \tab 10 \tab 32 \tab 23 \tab 8 \tab 9 \tab 8 \tab 14 \tab 6 \tab 5 \tab 9 \tab 6 \tab 51 \tab 16 \tab 1 \tab 6 \tab 3 \tab 4 \tab 6 \tab 5 \tab 3 \tab 4 \tab 3 \tab 37 \tab 11 \tab 5 \tab 4 \tab 4 \tab 5 \tab 4 \tab 4 \tab 6 \tab 1 \tab 3 \tab 36 \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 39 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 28 \tab 13 \tab 1470 \cr
   Total \tab 42 \tab 134 \tab 31 \tab 46 \tab 28 \tab 28 \tab 22 \tab 16 \tab 35 \tab 21 \tab 19 \tab 26 \tab 21 \tab 75 \tab 44 \tab 26 \tab 27 \tab 17 \tab 21 \tab 31 \tab 16 \tab 25 \tab 18 \tab 20 \tab 52 \tab 33 \tab 17 \tab 11 \tab 31 \tab 15 \tab 26 \tab 17 \tab 10 \tab 13 \tab 10 \tab 32 \tab 23 \tab 8 \tab 9 \tab 8 \tab 14 \tab 6 \tab 5 \tab 9 \tab 6 \tab 51 \tab 16 \tab 1 \tab 6 \tab 3 \tab 4 \tab 6 \tab 5 \tab 3 \tab 4 \tab 3 \tab 37 \tab 11 \tab 5 \tab 4 \tab 4 \tab 5 \tab 4 \tab 4 \tab 6 \tab 1 \tab 3 \tab 36 \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 39 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 28 \tab 13 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
