% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{livecom1}
\alias{livecom1}
\title{How long have you lived in the community?}
\source{
General Social Survey https://gss.norc.org
}
\description{
livecom1
}
\details{
Question How long have you lived in the community where you live now?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4866/vshow}{GSS Data Explorer page for \code{livecom1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 57 \tab 58 \tab 6 \tab 60 \tab 61 \tab 62 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 7 \tab 70 \tab 71 \tab 72 \tab 76 \tab 79 \tab 8 \tab 80 \tab 85 \tab 86 \tab 9 \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1537 \tab 185 \tab 52 \tab 16 \tab 25 \tab 14 \tab 16 \tab 35 \tab 20 \tab 15 \tab 29 \tab 16 \tab 88 \tab 50 \tab 17 \tab 13 \tab 21 \tab 20 \tab 29 \tab 13 \tab 11 \tab 13 \tab 11 \tab 71 \tab 44 \tab 7 \tab 14 \tab 13 \tab 6 \tab 14 \tab 8 \tab 11 \tab 11 \tab 4 \tab 59 \tab 23 \tab 5 \tab 8 \tab 8 \tab 5 \tab 10 \tab 5 \tab 3 \tab 8 \tab 5 \tab 60 \tab 12 \tab 4 \tab 4 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 45 \tab 7 \tab 5 \tab 4 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 30 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 34 \tab 2 \tab 3 \tab 1 \tab 27 \tab 3 \tab 46 \tab - \tab 2904 \cr
   Total \tab 1537 \tab 185 \tab 52 \tab 16 \tab 25 \tab 14 \tab 16 \tab 35 \tab 20 \tab 15 \tab 29 \tab 16 \tab 88 \tab 50 \tab 17 \tab 13 \tab 21 \tab 20 \tab 29 \tab 13 \tab 11 \tab 13 \tab 11 \tab 71 \tab 44 \tab 7 \tab 14 \tab 13 \tab 6 \tab 14 \tab 8 \tab 11 \tab 11 \tab 4 \tab 59 \tab 23 \tab 5 \tab 8 \tab 8 \tab 5 \tab 10 \tab 5 \tab 3 \tab 8 \tab 5 \tab 60 \tab 12 \tab 4 \tab 4 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 45 \tab 7 \tab 5 \tab 4 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 30 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 34 \tab 2 \tab 3 \tab 1 \tab 27 \tab 3 \tab 46 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amproud1}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{belikeus}},
\code{\link{biasnews}},
\code{\link{bizmoney}},
\code{\link{brneffrt}},
\code{\link{citizen}},
\code{\link{citworld}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{decsorgs}},
\code{\link{ethclose}},
\code{\link{experts}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{freetrde}},
\code{\link{fulldem}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immassim}},
\code{\link{immcosts}},
\code{\link{immcrime1}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{immideas}},
\code{\link{immideas1}},
\code{\link{immimp}},
\code{\link{immjobs1}},
\code{\link{immrghts}},
\code{\link{immwlfare}},
\code{\link{imports}},
\code{\link{incgap1}},
\code{\link{internet}},
\code{\link{intlincs}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{lessprd}},
\code{\link{letin1}},
\code{\link{letin1a}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{refugees}},
\code{\link{res161}},
\code{\link{shortcom}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{wrldgovt}}

Other Residence: 
\code{\link{abroad}},
\code{\link{agecmeus}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{comamind}},
\code{\link{comtype}},
\code{\link{comyear}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwell5}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{geomobil}},
\code{\link{hspasn10}},
\code{\link{liveasns}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{livehsps}},
\code{\link{loclived}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{res16}},
\code{\link{res161}},
\code{\link{res2006}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{ISSP National Identity}
\concept{Residence}
\keyword{variable}
