% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{livehome}
\alias{livehome}
\title{Length at present dwelling}
\source{
General Social Survey https://gss.norc.org
}
\description{
livehome
}
\details{
Question How long have you lived in your present home, that is, the address where you live now? PLEASE FILL IN THE NUMBER OF YEARS IN THE SPACE BELOW. IF LESS THAN ONE YEAR,  MAKE  A CHECK ON THE LINE NEXT TO THAT CHOICE.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4182/vshow}{GSS Data Explorer page for \code{livehome}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 43 \tab 44 \tab 45 \tab 46 \tab 48 \tab 49 \tab 5 \tab 50 \tab 54 \tab 55 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 67 \tab 7 \tab 70 \tab 75 \tab 8 \tab 80 \tab 83 \tab 9 \tab no answer \tab not available in this year \tab Total \cr
   1986 \tab 42 \tab 261 \tab 74 \tab 54 \tab 27 \tab 32 \tab 19 \tab 21 \tab 37 \tab 23 \tab 17 \tab 22 \tab 21 \tab 134 \tab 37 \tab 13 \tab 16 \tab 17 \tab 10 \tab 17 \tab 17 \tab 7 \tab 10 \tab 10 \tab 87 \tab 14 \tab 10 \tab 5 \tab 13 \tab 4 \tab 11 \tab 4 \tab 4 \tab 5 \tab 4 \tab 56 \tab 5 \tab 3 \tab 1 \tab 4 \tab 5 \tab 3 \tab 2 \tab 1 \tab 68 \tab 5 \tab 1 \tab 2 \tab 57 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 45 \tab 2 \tab 1 \tab 58 \tab 1 \tab 1 \tab 34 \tab 7 \tab - \tab 1470 \cr
   Total \tab 42 \tab 261 \tab 74 \tab 54 \tab 27 \tab 32 \tab 19 \tab 21 \tab 37 \tab 23 \tab 17 \tab 22 \tab 21 \tab 134 \tab 37 \tab 13 \tab 16 \tab 17 \tab 10 \tab 17 \tab 17 \tab 7 \tab 10 \tab 10 \tab 87 \tab 14 \tab 10 \tab 5 \tab 13 \tab 4 \tab 11 \tab 4 \tab 4 \tab 5 \tab 4 \tab 56 \tab 5 \tab 3 \tab 1 \tab 4 \tab 5 \tab 3 \tab 2 \tab 1 \tab 68 \tab 5 \tab 1 \tab 2 \tab 57 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 45 \tab 2 \tab 1 \tab 58 \tab 1 \tab 1 \tab 34 \tab 7 \tab 0 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{bestfrd}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{comyear}},
\code{\link{cousins}},
\code{\link{cowrkfrd}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{hapunhap}},
\code{\link{helpfrds}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{inlawnum}},
\code{\link{kidcall}},
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidpars}},
\code{\link{kidvisit}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{livecom}},
\code{\link{localgvt}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{malive}},
\code{\link{marcohab}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{neifrd}},
\code{\link{niecenep}},
\code{\link{othfrd}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{palive}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{peocntct}},
\code{\link{posslq}},
\code{\link{posslqy}},
\code{\link{relcall}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{talkedto}},
\code{\link{trustpeo}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{usefrds}},
\code{\link{wantbest}}

Other Residence: 
\code{\link{abroad}},
\code{\link{agecmeus}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{comamind}},
\code{\link{comtype}},
\code{\link{comyear}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwell5}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{geomobil}},
\code{\link{hspasn10}},
\code{\link{liveasns}},
\code{\link{livecom}},
\code{\link{livecom1}},
\code{\link{livehsps}},
\code{\link{loclived}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{res16}},
\code{\link{res161}},
\code{\link{res2006}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{Residence}
\keyword{variable}
