% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{livewhts}
\alias{livewhts}
\title{R favors living in half white neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
livewhts
}
\details{
Question Now I'm going to ask you about different types of contact with various groups of people. In each situation would you please tell me whether you would be very much in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very much opposed to it happening?
G. Living in a neighborhood where half of your neighbors were whites?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1220/vshow}{GSS Data Explorer page for \code{livewhts}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2000 \tab 1419 \tab 412 \tab 294 \tab 537 \tab 71 \tab 23 \tab 13 \tab 48 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 257 \tab 212 \tab 365 \tab 48 \tab 19 \tab 7 \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 167 \tab 201 \tab 460 \tab 52 \tab 17 \tab 7 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 429 \tab 446 \tab 972 \tab 104 \tab 24 \tab 13 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 258 \tab 310 \tab 661 \tab 61 \tab 23 \tab 13 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 308 \tab 354 \tab 686 \tab 47 \tab 16 \tab 13 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 267 \tab 315 \tab 642 \tab 40 \tab 22 \tab 14 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 322 \tab 371 \tab 878 \tab 65 \tab 23 \tab 12 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 368 \tab 453 \tab 945 \tab 78 \tab 28 \tab 11 \tab 5 \tab - \tab 2867 \cr
   Total \tab 11522 \tab 2788 \tab 2956 \tab 6146 \tab 566 \tab 195 \tab 103 \tab 74 \tab 0 \tab 24350 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Whites: 
\code{\link{acqfmwht}},
\code{\link{acqnhwht}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkwht}},
\code{\link{closewht}},
\code{\link{comwht}},
\code{\link{fairwhts}},
\code{\link{famwhts}},
\code{\link{farewhts}},
\code{\link{feelwhts}},
\code{\link{influwht}},
\code{\link{intlwhts}},
\code{\link{knwwht}},
\code{\link{liveno}},
\code{\link{marno}},
\code{\link{marwht}},
\code{\link{patrwhts}},
\code{\link{trtwhite}},
\code{\link{uswht}},
\code{\link{violwhts}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtsdiff}},
\code{\link{whtwrk}},
\code{\link{wlthwhts}},
\code{\link{workwhts}}
}
\concept{Intergroup Relations}
\concept{Whites}
\keyword{variable}
