% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{lkelyvot}
\alias{lkelyvot}
\title{Likelihood to vote in 2024 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
lkelyvot
}
\details{
Question How likely are you to vote in the November 2024 presidential election?
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely likely
\item \verb{[2]} very likely
\item \verb{[3]} somewhat likely
\item \verb{[4]} not very likely
\item \verb{[5]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8146/vshow}{GSS Data Explorer page for \code{lkelyvot}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab extremely likely \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1807 \tab 406 \tab 303 \tab 135 \tab 297 \tab 17 \tab 321 \tab 12 \tab 11 \tab 3309 \cr
   Total \tab 0 \tab 1807 \tab 406 \tab 303 \tab 135 \tab 297 \tab 17 \tab 321 \tab 12 \tab 11 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{econstat}},
\code{\link{polintrst}},
\code{\link{whovote24a}}
}
\concept{ANES}
\keyword{variable}
