% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_197.R
\name{loanedto}
\alias{loanedto}
\title{Oto whom did you lend this money?}
\source{
General Social Survey https://gss.norc.org
}
\description{
loanedto
}
\details{
Question Think of the most recent loan of this type. To whom did you lend this money?
}
\section{Values}{

\itemize{
\item \verb{[1]} parents of grandparents
\item \verb{[2]} children
\item \verb{[3]} other relative
\item \verb{[4]} friend
\item \verb{[5]} co-worker or bus. associate
\item \verb{[6]} other specify
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1856/vshow}{GSS Data Explorer page for \code{loanedto}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab parents of grandparents \tab children \tab other relative \tab friend \tab co-worker or bus. associate \tab other specify \tab Total \cr
   1996 \tab 2620 \tab 15 \tab 66 \tab 68 \tab 103 \tab 26 \tab 6 \tab 2904 \cr
   Total \tab 2620 \tab 15 \tab 66 \tab 68 \tab 103 \tab 26 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{loaned}},
\code{\link{sellauto}},
\code{\link{sellhome}},
\code{\link{transoth}}
}
\concept{Markets}
\keyword{variable}
