% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{loanitem}
\alias{loanitem}
\title{R has let someone borrow a item of some value}
\source{
General Social Survey https://gss.norc.org
}
\description{
loanitem
}
\details{
Question During the past 12 months, how often have you done each of the following things:
K. Let someone you didn''t know well borrow a item of some value like dishes or tools.
}
\section{Values}{

\itemize{
\item \code{1} more than once a week
\item \code{2} once a week
\item \code{3} once a month
\item \code{4} at least 2 or 3 times in the past year
\item \code{5} once in the past year
\item \code{6} not at all in the past year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2012 \tab A/C/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2887/vshow}{GSS Data Explorer page for \code{loanitem}}.
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selffrst}},
\code{\link{selfless}}

Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selffrst}},
\code{\link{selfless}}
}
\concept{Altruism}
\keyword{variable}
