% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{lobbygrp}
\alias{lobbygrp}
\title{Has r contacted government for group}
\source{
General Social Survey https://gss.norc.org
}
\description{
lobbygrp
}
\details{
Question Here is a list of things that members of organizations may participate in. Could you tell me whether or not you have ever done any of these things as a member of (GROUP IN MEMFRAT-MEMNUM OR 332)? (READ EACH ITEM. CIRCLE A NUMBER FOR EACH ACTIVITY).
F. Contacted government officials on behalf of the organization
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1020/vshow}{GSS Data Explorer page for \code{lobbygrp}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab 606 \tab 220 \tab 953 \tab 40 \tab 1819 \cr
   Total \tab 606 \tab 220 \tab 953 \tab 40 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
