% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{loccare}
\alias{loccare}
\title{Would local officials listen to problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
loccare
}
\details{
Question If you had some complaint about a local government activity and took that complaint to a member of the local government council, would you expect him or her to pay a lot of attention to what you say, some attention, very little attention, or none at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} lot of attention
\item \verb{[2]} some
\item \verb{[3]} very little
\item \verb{[4]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1026/vshow}{GSS Data Explorer page for \code{loccare}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab lot of attention \tab some \tab very little \tab none \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 487 \tab 859 \tab 307 \tab 121 \tab 35 \tab 10 \tab 1819 \cr
   Total \tab 0 \tab 487 \tab 859 \tab 307 \tab 121 \tab 35 \tab 10 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
