% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{locinflu}
\alias{locinflu}
\title{How much influence r feels on local gov}
\source{
General Social Survey https://gss.norc.org
}
\description{
locinflu
}
\details{
Question How much influence do you think people like you can have over local government decisions -- a lot, a moderate amount, a little, or none at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} a lot
\item \verb{[2]} moderate amount
\item \verb{[3]} a little
\item \verb{[4]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1023/vshow}{GSS Data Explorer page for \code{locinflu}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a lot \tab moderate amount \tab a little \tab none \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 252 \tab 585 \tab 650 \tab 292 \tab 33 \tab 7 \tab 1819 \cr
   Total \tab 0 \tab 252 \tab 585 \tab 650 \tab 292 \tab 33 \tab 7 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
