% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{loclobby}
\alias{loclobby}
\title{Has r lobbied local officials}
\source{
General Social Survey https://gss.norc.org
}
\description{
loclobby
}
\details{
Question Have you ever personally gone to see, or spoken to, or written to -- some member of local government or some other person of influence in the community about some needs or problems?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1031/vshow}{GSS Data Explorer page for \code{loclobby}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab - \tab 594 \tab 1216 \tab 9 \tab 1819 \cr
   Total \tab 0 \tab 594 \tab 1216 \tab 9 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
