% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{locprob}
\alias{locprob}
\title{Has r helped to solve local problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
locprob
}
\details{
Question Have you ever worked with others in this community to try to solve some community problems?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1024/vshow}{GSS Data Explorer page for \code{locprob}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}

Other Political: 
\code{\link{intpol}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Political}
\concept{Socio-political participation}
\keyword{variable}
