% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{loctrust}
\alias{loctrust}
\title{How much trust does r have in local gov}
\source{
General Social Survey https://gss.norc.org
}
\description{
loctrust
}
\details{
Question How much of the time do you think you can trust the local government here in (MENTION NAME OF LOCAL GOVERNMENT UNIT) to do what is right -- just about always, most of the time, only some of the time, or almost ever?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} most of time
\item \code{3} some of time
\item \code{4} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1022/vshow}{GSS Data Explorer page for \code{loctrust}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab most of time \tab some of time \tab almost never \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 196 \tab 750 \tab 653 \tab 130 \tab 78 \tab 12 \tab 1819 \cr
   Total \tab 0 \tab 196 \tab 750 \tab 653 \tab 130 \tab 78 \tab 12 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
