% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{locvote}
\alias{locvote}
\title{How often does r vote in local elections}
\source{
General Social Survey https://gss.norc.org
}
\description{
locvote
}
\details{
Question What about local elections -- do you always vote in those, do you sometimes miss one, or do you rarely vote, or do you never vote?
}
\section{Values}{

\itemize{
\item \code{1} vote in all
\item \code{2} sometimes miss
\item \code{3} rarely vote
\item \code{4} never vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1039/vshow}{GSS Data Explorer page for \code{locvote}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}

Other Political: 
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Political}
\concept{Socio-political participation}
\keyword{variable}
