% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{locyrs}
\alias{locyrs}
\title{How long have worked at this location}
\source{
General Social Survey https://gss.norc.org
}
\description{
locyrs
}
\details{
Question How long have you worked at this location?  Include all full-time positions you have had at this location.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2146/vshow}{GSS Data Explorer page for \code{locyrs}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 0.08 \tab 0.17 \tab 0.25 \tab 0.33 \tab 0.42 \tab 0.5 \tab 0.58 \tab 0.67 \tab 0.75 \tab 0.83 \tab 0.92 \tab 1 \tab 1.08 \tab 1.17 \tab 1.25 \tab 1.33 \tab 1.42 \tab 1.5 \tab 1.58 \tab 1.67 \tab 1.75 \tab 1.83 \tab 10 \tab 10.2 \tab 10.3 \tab 10.5 \tab 10.7 \tab 10.8 \tab 11 \tab 11.1 \tab 11.6 \tab 11.9 \tab 12 \tab 12.5 \tab 12.8 \tab 12.9 \tab 13 \tab 13.1 \tab 14 \tab 14.2 \tab 14.3 \tab 15 \tab 15.1 \tab 15.4 \tab 15.6 \tab 16 \tab 17 \tab 17.2 \tab 17.5 \tab 18 \tab 18.1 \tab 18.8 \tab 19 \tab 19.4 \tab 2 \tab 2.17 \tab 2.25 \tab 2.33 \tab 2.42 \tab 2.5 \tab 2.58 \tab 2.67 \tab 2.75 \tab 2.83 \tab 2.92 \tab 20 \tab 20.2 \tab 21 \tab 21.4 \tab 21.5 \tab 21.8 \tab 22 \tab 22.1 \tab 23 \tab 24 \tab 24.4 \tab 25 \tab 26 \tab 27 \tab 28 \tab 28.3 \tab 28.9 \tab 29 \tab 29.3 \tab 3 \tab 3.08 \tab 3.17 \tab 3.25 \tab 3.33 \tab 3.5 \tab 3.58 \tab 3.67 \tab 3.83 \tab 30 \tab 30.4 \tab 31.1 \tab 32 \tab 33 \tab 34 \tab 36 \tab 4 \tab 4.17 \tab 4.25 \tab 4.33 \tab 4.42 \tab 4.5 \tab 4.58 \tab 4.67 \tab 4.75 \tab 4.83 \tab 40 \tab 45 \tab 5 \tab 5.08 \tab 5.17 \tab 5.25 \tab 5.42 \tab 5.5 \tab 5.67 \tab 6 \tab 6.17 \tab 6.25 \tab 6.33 \tab 6.5 \tab 6.75 \tab 6.92 \tab 7 \tab 7.17 \tab 7.25 \tab 7.33 \tab 7.5 \tab 7.58 \tab 7.92 \tab 8 \tab 8.5 \tab 8.67 \tab 9 \tab 9.08 \tab 9.25 \tab 9.5 \tab 9.67 \tab 9.75 \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 1455 \tab 53 \tab 24 \tab 18 \tab 32 \tab 19 \tab 14 \tab 48 \tab 17 \tab 28 \tab 12 \tab 20 \tab 2 \tab 90 \tab 9 \tab 10 \tab 3 \tab 2 \tab 6 \tab 32 \tab 2 \tab 4 \tab 3 \tab 2 \tab 49 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 18 \tab 1 \tab 1 \tab 1 \tab 22 \tab 1 \tab 1 \tab 1 \tab 16 \tab 1 \tab 19 \tab 1 \tab 1 \tab 25 \tab 1 \tab 2 \tab 1 \tab 2 \tab 10 \tab 1 \tab 1 \tab 13 \tab 3 \tab 1 \tab 11 \tab 1 \tab 90 \tab 2 \tab 3 \tab 4 \tab 4 \tab 20 \tab 5 \tab 2 \tab 2 \tab 1 \tab 2 \tab 23 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 5 \tab 4 \tab 1 \tab 9 \tab 3 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 70 \tab 3 \tab 3 \tab 4 \tab 4 \tab 16 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 54 \tab 3 \tab 2 \tab 2 \tab 2 \tab 7 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 66 \tab 1 \tab 5 \tab 1 \tab 1 \tab 7 \tab 1 \tab 42 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 47 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 34 \tab 2 \tab 1 \tab 33 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 37 \tab - \tab 2832 \cr
   Total \tab 1455 \tab 53 \tab 24 \tab 18 \tab 32 \tab 19 \tab 14 \tab 48 \tab 17 \tab 28 \tab 12 \tab 20 \tab 2 \tab 90 \tab 9 \tab 10 \tab 3 \tab 2 \tab 6 \tab 32 \tab 2 \tab 4 \tab 3 \tab 2 \tab 49 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 18 \tab 1 \tab 1 \tab 1 \tab 22 \tab 1 \tab 1 \tab 1 \tab 16 \tab 1 \tab 19 \tab 1 \tab 1 \tab 25 \tab 1 \tab 2 \tab 1 \tab 2 \tab 10 \tab 1 \tab 1 \tab 13 \tab 3 \tab 1 \tab 11 \tab 1 \tab 90 \tab 2 \tab 3 \tab 4 \tab 4 \tab 20 \tab 5 \tab 2 \tab 2 \tab 1 \tab 2 \tab 23 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 5 \tab 4 \tab 1 \tab 9 \tab 3 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 70 \tab 3 \tab 3 \tab 4 \tab 4 \tab 16 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 54 \tab 3 \tab 2 \tab 2 \tab 2 \tab 7 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 66 \tab 1 \tab 5 \tab 1 \tab 1 \tab 7 \tab 1 \tab 42 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 47 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 34 \tab 2 \tab 1 \tab 33 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 37 \tab 0 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{daily}},
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{monthly}},
\code{\link{numemp}},
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{trainsch}},
\code{\link{trusting}},
\code{\link{typeorg}},
\code{\link{weekly}},
\code{\link{wojob}},
\code{\link{wojobyrs}},
\code{\link{wrkoth}},
\code{\link{yearly}}

Other Job: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{actupset}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{baseofop}},
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{bosswrks}},
\code{\link{chngeorg}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compperf}},
\code{\link{condemnd}},
\code{\link{cowrkers}},
\code{\link{cowrkhlp}},
\code{\link{cowrkhrd}},
\code{\link{cowrkint}},
\code{\link{cowrkrel}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{dailywrk}},
\code{\link{danger}},
\code{\link{datefrst}},
\code{\link{denyrais}},
\code{\link{deptperf}},
\code{\link{difstand}},
\code{\link{discaff}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{discaffy}},
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{donothng}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{exptrain}},
\code{\link{extr2001}},
\code{\link{extr2005}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{famintjb}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{fehire}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findnwjb}},
\code{\link{findout}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{formltrn}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{fringeok}},
\code{\link{ftwork}},
\code{\link{ftwork1}},
\code{\link{gdjobsec}},
\code{\link{genderwk}},
\code{\link{handmove}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{haveinfo}},
\code{\link{helporg}},
\code{\link{hiinc}},
\code{\link{hlpequip}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hlthtype}},
\code{\link{hotargus}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{hrsmoney}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{imatter}},
\code{\link{inclstjb}},
\code{\link{indperf}},
\code{\link{indusoth}},
\code{\link{intjob}},
\code{\link{intlcowk}},
\code{\link{iwrkhard}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobaff}},
\code{\link{jobcntct}},
\code{\link{jobcntrl}},
\code{\link{jobcowrk}},
\code{\link{jobeasy}},
\code{\link{jobfrnds}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobinfo}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobpay}},
\code{\link{jobplace}},
\code{\link{jobpromo}},
\code{\link{jobpub}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobsecok}},
\code{\link{jobskill}},
\code{\link{jobtvrad}},
\code{\link{jobvshme}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{knowschd}},
\code{\link{knowwhat}},
\code{\link{lackinfo}},
\code{\link{lastjob}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{learning}},
\code{\link{learnjob}},
\code{\link{learnnew}},
\code{\link{leavejb}},
\code{\link{leavejob}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{localnum}},
\code{\link{lookaway}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{lotofsay}},
\code{\link{lottosay}},
\code{\link{manvsemp}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{misswork}},
\code{\link{monitred}},
\code{\link{monthly}},
\code{\link{moonlite}},
\code{\link{moretrde}},
\code{\link{mykids}},
\code{\link{myskills}},
\code{\link{negjob5}},
\code{\link{newjob}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{ntwkhard}},
\code{\link{numemp}},
\code{\link{numemps5}},
\code{\link{numorg}},
\code{\link{numown}},
\code{\link{numsites}},
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{offsup}},
\code{\link{opdevel}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{orgsize}},
\code{\link{othcredt}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{otherwrk}},
\code{\link{othhrs}},
\code{\link{othplan}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{overwork}},
\code{\link{owninc}},
\code{\link{ownstock}},
\code{\link{paidemps}},
\code{\link{parleave}},
\code{\link{partorg}},
\code{\link{partteam}},
\code{\link{pastpay}},
\code{\link{pawrkslf}},
\code{\link{paychnge}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{perspace}},
\code{\link{phyeffrt}},
\code{\link{physharm}},
\code{\link{physwrk}},
\code{\link{placewrk}},
\code{\link{planwrk}},
\code{\link{privgovt}},
\code{\link{prodctiv}},
\code{\link{promotn}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{proudorg}},
\code{\link{putdown}},
\code{\link{rageemp}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{rdiscaff}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{reliedon}},
\code{\link{replaceu}},
\code{\link{respect}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rimmdisc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpaysame}},
\code{\link{rpromotn}},
\code{\link{rptprobs}},
\code{\link{rsecjob}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{rwrkindp}},
\code{\link{ryrsfirm}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{samevals}},
\code{\link{samework}},
\code{\link{satjob}},
\code{\link{satjob1}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{schoolng}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{secondwk}},
\code{\link{seecowrk}},
\code{\link{sethours}},
\code{\link{setthngs}},
\code{\link{shout}},
\code{\link{siteyrs}},
\code{\link{skipwork}},
\code{\link{slsmnshp}},
\code{\link{smallbig}},
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{splocnum}},
\code{\link{spvtrfair}},
\code{\link{spwrkslf}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{stress}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supervis}},
\code{\link{suphelp}},
\code{\link{supnum}},
\code{\link{supothrs}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{teamsafe}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timeoff}},
\code{\link{toldwork}},
\code{\link{toofewwk}},
\code{\link{totalnum}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{training}},
\code{\link{trainops}},
\code{\link{trainsch}},
\code{\link{trdestck}},
\code{\link{treatres}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{typeorg}},
\code{\link{unhlthy}},
\code{\link{unmanrel}},
\code{\link{unpleznt}},
\code{\link{useskill}},
\code{\link{usetech}},
\code{\link{wantjob}},
\code{\link{wantjob1}},
\code{\link{weekly}},
\code{\link{wherewrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkbhvrs}},
\code{\link{wkbonus}},
\code{\link{wkcomptr}},
\code{\link{wkcontct}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkpraise}},
\code{\link{wksmooth}},
\code{\link{wksub}},
\code{\link{wksup}},
\code{\link{wktenure}},
\code{\link{wktopsat}},
\code{\link{wocomptr}},
\code{\link{work3yrs}},
\code{\link{workdiff}},
\code{\link{workdy}},
\code{\link{workfast}},
\code{\link{workless}},
\code{\link{worryjob}},
\code{\link{wrkindp}},
\code{\link{wrkoth}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrktime}},
\code{\link{xhaustn}},
\code{\link{yearbiz}},
\code{\link{yearly}},
\code{\link{yearsjob}},
\code{\link{yearval}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Job}
\concept{Jobs}
\keyword{variable}
