% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{lookaway}
\alias{lookaway}
\title{Ppl look the other way when others are threatened}
\source{
General Social Survey https://gss.norc.org
}
\description{
lookaway
}
\details{
Question N. In my workplace, people "look the other way" when others are threatened, intimidated, or put down.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3011/vshow}{GSS Data Explorer page for \code{lookaway}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 57 \tab 243 \tab 801 \tab 683 \tab 39 \tab 44 \tab 2812 \cr
   2012 \tab 795 \tab 51 \tab 180 \tab 465 \tab 407 \tab 40 \tab 36 \tab 1974 \cr
   Total \tab 1740 \tab 108 \tab 423 \tab 1266 \tab 1090 \tab 79 \tab 80 \tab 4786 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
