% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{lookjob}
\alias{lookjob}
\title{Are you looking for a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
lookjob
}
\details{
Question Are you currently looking for a job?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4598/vshow}{GSS Data Explorer page for \code{lookjob}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1998 \tab 2460 \tab 65 \tab 276 \tab 31 \tab - \tab 2832 \cr
   2006 \tab 4008 \tab 85 \tab 413 \tab 3 \tab 1 \tab 4510 \cr
   2016 \tab 2533 \tab 94 \tab 216 \tab 23 \tab 1 \tab 2867 \cr
   Total \tab 9001 \tab 244 \tab 905 \tab 57 \tab 2 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
