% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{losejb12}
\alias{losejb12}
\title{How likely r will lose job or laid off in next 1 yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
losejb12
}
\details{
Question Thinking now about the next 12 months, how likely do you think it is that you will lose your job or be laid off for reasons other than for your performance? Would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} fairly likely
\item \verb{[3]} not too likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3879/vshow}{GSS Data Explorer page for \code{losejb12}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab fairly likely \tab not too likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 28 \tab 47 \tab 264 \tab 495 \tab 12 \tab 12 \tab 2023 \cr
   Total \tab 1165 \tab 28 \tab 47 \tab 264 \tab 495 \tab 12 \tab 12 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\keyword{variable}
