% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{lotofsay}
\alias{lotofsay}
\title{R has lot of say in job}
\source{
General Social Survey https://gss.norc.org
}
\description{
lotofsay
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
D. I have a lot of say about what happens on my job
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2777/vshow}{GSS Data Explorer page for \code{lotofsay}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 545 \tab 741 \tab 385 \tab 101 \tab 11 \tab 13 \tab 2765 \cr
   2006 \tab 2776 \tab 504 \tab 675 \tab 416 \tab 117 \tab 10 \tab 12 \tab 4510 \cr
   Total \tab 3745 \tab 1049 \tab 1416 \tab 801 \tab 218 \tab 21 \tab 25 \tab 7275 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
