% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_230.R
\name{lotr1}
\alias{lotr1}
\title{In uncertain times i usually expect best}
\source{
General Social Survey https://gss.norc.org
}
\description{
lotr1
}
\details{
Question (HAND CARD G1)                       
Please answer the following questions about yourself by indicating the extent of your agreement using the following scale: Strongly disagree, Disagree, Neutral, Agree, and Strongly agree.
Be as honest as you can throughout, and try not to let your responses to one question influence your response to other questions. There are no right or wrong answers.  Do you agree or disagree that...
In uncertain times, I usually expect the best.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly disagree
\item \verb{[2]} disagree
\item \verb{[3]} neutral
\item \verb{[4]} agree
\item \verb{[5]} strongly agree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6197/vshow}{GSS Data Explorer page for \code{lotr1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly disagree \tab disagree \tab neutral \tab agree \tab strongly agree \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1409 \tab 70 \tab 196 \tab 271 \tab 683 \tab 227 \tab 2 \tab 9 \tab 2867 \cr
   Total \tab 1409 \tab 70 \tab 196 \tab 271 \tab 683 \tab 227 \tab 2 \tab 9 \tab 2867 \cr
}
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope1}},
\code{\link{hope2}},
\code{\link{hope3}},
\code{\link{hope4}},
\code{\link{lotr2}},
\code{\link{lotr3}},
\code{\link{lotr4}},
\code{\link{lotr5}},
\code{\link{lotr6}}
}
\concept{Hope and Optimism}
\keyword{variable}
