% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_230.R
\name{lotr6}
\alias{lotr6}
\title{I expect more good things to happen to me than bad}
\source{
General Social Survey https://gss.norc.org
}
\description{
lotr6
}
\details{
Question (HAND CARD G1)
(Do you agree or disagree that . . .)
Overall, I expect more good things to happen to me than bad.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly disagree
\item \verb{[2]} disagree
\item \verb{[3]} neutral
\item \verb{[4]} agree
\item \verb{[5]} strongly agree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6202/vshow}{GSS Data Explorer page for \code{lotr6}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly disagree \tab disagree \tab neutral \tab agree \tab strongly agree \tab no answer \tab Total \cr
   2016 \tab 1409 \tab 39 \tab 90 \tab 196 \tab 843 \tab 280 \tab 10 \tab 2867 \cr
   Total \tab 1409 \tab 39 \tab 90 \tab 196 \tab 843 \tab 280 \tab 10 \tab 2867 \cr
}
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope1}},
\code{\link{hope2}},
\code{\link{hope3}},
\code{\link{hope4}},
\code{\link{lotr1}},
\code{\link{lotr2}},
\code{\link{lotr3}},
\code{\link{lotr4}},
\code{\link{lotr5}}
}
\concept{Hope and Optimism}
\keyword{variable}
