% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{lowpay}
\alias{lowpay}
\title{I would accept job for lower pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
lowpay
}
\details{
Question (HAND CARD D22)
(To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing...)
To accept a position with lower pay.
(Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6344/vshow}{GSS Data Explorer page for \code{lowpay}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2533 \tab 32 \tab 116 \tab 58 \tab 68 \tab 33 \tab 4 \tab 23 \tab 2867 \cr
   Total \tab 2533 \tab 32 \tab 116 \tab 58 \tab 68 \tab 33 \tab 4 \tab 23 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
