% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{maattend}
\alias{maattend}
\title{How often ma attended religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
maattend
}
\details{
Question A. When you were growing up, how often did your mother (MOTHER SUBSTITUTE) attend religious services?
}
\section{Values}{

\itemize{
\item \verb{[0]} never
\item \verb{[1]} lt once a year
\item \verb{[2]} once a year
\item \verb{[3]} sevrl times a yr
\item \verb{[4]} once a month
\item \verb{[5]} 2-3x a month
\item \verb{[6]} nrly every week
\item \verb{[7]} every week
\item \verb{[8]} more thn once wk
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1983, 1984, 1985, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/295/vshow}{GSS Data Explorer page for \code{maattend}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab never \tab lt once a year \tab once a year \tab sevrl times a yr \tab once a month \tab 2-3x a month \tab nrly every week \tab every week \tab more thn once wk \tab don't know \tab Total \cr
   1983 \tab 48 \tab 134 \tab 51 \tab 104 \tab 116 \tab 105 \tab 126 \tab 146 \tab 592 \tab 159 \tab 18 \tab 1599 \cr
   1984 \tab 68 \tab 113 \tab 44 \tab 87 \tab 122 \tab 75 \tab 93 \tab 97 \tab 604 \tab 146 \tab 24 \tab 1473 \cr
   1985 \tab 67 \tab 142 \tab 43 \tab 106 \tab 96 \tab 73 \tab 101 \tab 100 \tab 639 \tab 157 \tab 10 \tab 1534 \cr
   1986 \tab 47 \tab 129 \tab 32 \tab 98 \tab 108 \tab 73 \tab 114 \tab 109 \tab 592 \tab 152 \tab 16 \tab 1470 \cr
   1988 \tab 48 \tab 125 \tab 33 \tab 91 \tab 124 \tab 95 \tab 115 \tab 155 \tab 547 \tab 130 \tab 18 \tab 1481 \cr
   1989 \tab 560 \tab 109 \tab 25 \tab 59 \tab 82 \tab 43 \tab 62 \tab 64 \tab 419 \tab 103 \tab 11 \tab 1537 \cr
   Total \tab 838 \tab 752 \tab 228 \tab 545 \tab 648 \tab 464 \tab 611 \tab 671 \tab 3393 \tab 847 \tab 97 \tab 9094 \cr
}
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jew}},
\code{\link{jewaj}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{spattend}}
}
\concept{Core}
\keyword{variable}
