% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{maborn}
\alias{maborn}
\title{Was r's mother (or youngest same-sex parent) born in this country?}
\source{
General Social Survey https://gss.norc.org
}
\description{
maborn
}
\details{
Question Was your mother born in this country?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7476/vshow}{GSS Data Explorer page for \code{maborn}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 48 \tab 3312 \tab 627 \tab 27 \tab 18 \tab - \tab 4032 \cr
   2022 \tab - \tab 2841 \tab 669 \tab 19 \tab 8 \tab 7 \tab 3544 \cr
   Total \tab 48 \tab 6153 \tab 1296 \tab 46 \tab 26 \tab 7 \tab 7576 \cr
}
}

\keyword{variable}
