% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{maboygrl}
\alias{maboygrl}
\title{Does mom's gene decide baby's sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
maboygrl
}
\details{
Question It is the mother’s gene that decides whether the baby is a boy or a girl.
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3470/vshow}{GSS Data Explorer page for \code{maboygrl}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1769 \tab 41 \tab 176 \tab 35 \tab 2 \tab 2023 \cr
   Total \tab 1769 \tab 41 \tab 176 \tab 35 \tab 2 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
