% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_457.R
\name{macall1}
\alias{macall1}
\title{How often does r contact mother by telephone}
\source{
General Social Survey https://gss.norc.org
}
\description{
macall1
}
\details{
Question How often do you have any other contact with your mother besides visiting, either by telephone, letter, fax or e-mail?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} at least several times a week
\item \verb{[3]} at least once a week
\item \verb{[4]} at least once a month
\item \verb{[5]} several times a year
\item \verb{[6]} less often
\item \verb{[7]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4195/vshow}{GSS Data Explorer page for \code{macall1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab daily \tab at least several times a week \tab at least once a week \tab at least once a month \tab several times a year \tab less often \tab never \tab no answer \tab Total \cr
   2002 \tab 2083 \tab 127 \tab 175 \tab 185 \tab 89 \tab 36 \tab 28 \tab 38 \tab 4 \tab 2765 \cr
   Total \tab 2083 \tab 127 \tab 175 \tab 185 \tab 89 \tab 36 \tab 28 \tab 38 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{cousins}},
\code{\link{kidcall}},
\code{\link{mavisit1}},
\code{\link{niecenep}},
\code{\link{pacall1}},
\code{\link{parslaw}},
\code{\link{pavisit1}},
\code{\link{sibinlaw}},
\code{\link{uncaunts}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
