% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_186.R
\name{madat11}
\alias{madat11}
\title{Angry at a neighbor}
\source{
General Social Survey https://gss.norc.org
}
\description{
madat11
}
\details{
Question Who were you angry, irritated or annoyed at? (IF RESPONDENT GIVES A NAME, PROBE WITH "and who is name... what relationship is he/she to you?" IF UNSURE ABOUT RELATIONSHIP AFTER RESPONDENT'S DESCRIPTION, READ THE FIELD CODE THAT YOU THINK IS CLOSEST -- e.g., "so that's a co-worker?") CIRCLE ALL THAT APPLY.
K. A neighbor
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1751/vshow}{GSS Data Explorer page for \code{madat11}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1779 \tab 25 \tab 1093 \tab 7 \tab 2904 \cr
   Total \tab 1779 \tab 25 \tab 1093 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{madat10}},
\code{\link{madat2}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat5}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}}
}
\concept{Emotions}
\keyword{variable}
