% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_187.R
\name{madat13}
\alias{madat13}
\title{Angry at a co-worker}
\source{
General Social Survey https://gss.norc.org
}
\description{
madat13
}
\details{
Question Who were you angry, irritated or annoyed at? (IF RESPONDENT GIVES A NAME, PROBE WITH "and who is name... what relationship is he/she to you?" IF UNSURE ABOUT RELATIONSHIP AFTER RESPONDENT'S DESCRIPTION, READ THE FIELD CODE THAT YOU THINK IS CLOSEST -- e.g., "so that's a co-worker?") CIRCLE ALL THAT APPLY.
M. A co-worker
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1753/vshow}{GSS Data Explorer page for \code{madat13}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1779 \tab 127 \tab 991 \tab 7 \tab 2904 \cr
   Total \tab 1779 \tab 127 \tab 991 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{madat12}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat19}},
\code{\link{madat20}},
\code{\link{madat21}}
}
\concept{Emotions}
\keyword{variable}
