% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_187.R
\name{madat19}
\alias{madat19}
\title{Angry at public figure}
\source{
General Social Survey https://gss.norc.org
}
\description{
madat19
}
\details{
Question Who were you angry, irritated or annoyed at? (IF RESPONDENT GIVES A NAME, PROBE WITH "and who is name... what relationship is he/she to you?" IF UNSURE ABOUT RELATIONSHIP AFTER RESPONDENT'S DESCRIPTION, READ THE FIELD CODE THAT YOU THINK IS CLOSEST -- e.g., "so that's a co-worker?") CIRCLE ALL THAT APPLY.
S. A public figure (politician, actor, newsmaker)
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1759/vshow}{GSS Data Explorer page for \code{madat19}}.
}

\seealso{
Other Emotions: 
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat20}},
\code{\link{madat21}}

Other Feelings: 
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat20}},
\code{\link{madat21}}
}
\concept{Emotions}
\concept{Feelings}
\keyword{variable}
