% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{madat23}
\alias{madat23}
\title{Angry at other}
\source{
General Social Survey https://gss.norc.org
}
\description{
madat23
}
\details{
Question Who were you angry, irritated or annoyed at? (IF RESPONDENT GIVES A NAME, PROBE WITH "and who is name... what relationship is he/she to you?" IF UNSURE ABOUT RELATIONSHIP AFTER RESPONDENT'S DESCRIPTION, READ THE FIELD CODE THAT YOU THINK IS CLOSEST -- e.g., "so that's a co-worker?") CIRCLE ALL THAT APPLY.
W. Other (SPECIFY)
}
\section{Values}{

\itemize{
\item \verb{[1]} ex-spouse
\item \verb{[2]} other ex-family
\item \verb{[3]} health care personnel
\item \verb{[4]} teacher
\item \verb{[5]} landlord
\item \verb{[6]} company
\item \verb{[7]} criminal
\item \verb{[8]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1763/vshow}{GSS Data Explorer page for \code{madat23}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab ex-spouse \tab other ex-family \tab health care personnel \tab teacher \tab landlord \tab company \tab criminal \tab other \tab no answer \tab Total \cr
   1996 \tab 2766 \tab 16 \tab 2 \tab 10 \tab 9 \tab 5 \tab 6 \tab 4 \tab 79 \tab 7 \tab 2904 \cr
   Total \tab 2766 \tab 16 \tab 2 \tab 10 \tab 9 \tab 5 \tab 6 \tab 4 \tab 79 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{howangry}},
\code{\link{madat22}},
\code{\link{thnkangr}}
}
\concept{Emotions}
\keyword{variable}
