% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_186.R
\name{madat5}
\alias{madat5}
\title{Angry at father}
\source{
General Social Survey https://gss.norc.org
}
\description{
madat5
}
\details{
Question Who were you angry, irritated or annoyed at? (IF RESPONDENT GIVES A NAME, PROBE WITH "and who is name... what relationship is he/she to you?" IF UNSURE ABOUT RELATIONSHIP AFTER RESPONDENT'S DESCRIPTION, READ THE FIELD CODE THAT YOU THINK IS CLOSEST -- e.g., "so that's a co-worker?") CIRCLE ALL THAT APPLY.
E. Father
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1745/vshow}{GSS Data Explorer page for \code{madat5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1779 \tab 5 \tab 1113 \tab 7 \tab 2904 \cr
   Total \tab 1779 \tab 5 \tab 1113 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{madat10}},
\code{\link{madat11}},
\code{\link{madat2}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}}
}
\concept{Emotions}
\keyword{variable}
