% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{madatgod}
\alias{madatgod}
\title{How often r feels angry at god}
\source{
General Social Survey https://gss.norc.org
}
\description{
madatgod
}
\details{
Question Here is another card with contrasting ideas. If you feel that
"My faith is completely free of doubts," you would place yourself at 1. If you feel "My faith is mixed with doubts," you would place yourself at 7. If you feel somewhere between these two, you would place yourself at 2, 3, 4, 5 or 6. Where would you place your feelings about your faith…
B. From Often felt agnry to Never felt angry at God
}
\section{Values}{

\itemize{
\item \verb{[1]} often angry at god
\item \verb{[7]} never angry at god
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1094/vshow}{GSS Data Explorer page for \code{madatgod}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab often angry at god \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab never angry at god \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 37 \tab 49 \tab 103 \tab 242 \tab 217 \tab 271 \tab 526 \tab 7 \tab 29 \tab 1481 \cr
   Total \tab 0 \tab 37 \tab 49 \tab 103 \tab 242 \tab 217 \tab 271 \tab 526 \tab 7 \tab 29 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
