% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{maden}
\alias{maden}
\title{What was mother's religious denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
maden
}
\details{
Question A. What specific denomination is that, if any? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \verb{[10]} am baptist asso
\item \verb{[11]} am bapt ch in usa
\item \verb{[12]} nat bapt conv of am
\item \verb{[13]} nat bapt conv usa
\item \verb{[14]} southern baptist
\item \verb{[15]} other baptists
\item \verb{[18]} baptist-dk which
\item \verb{[20]} afr meth episcopal
\item \verb{[21]} afr meth ep zion
\item \verb{[22]} united methodist
\item \verb{[23]} other methodist
\item \verb{[28]} methodist-dk which
\item \verb{[30]} am lutheran
\item \verb{[31]} luth ch in america
\item \verb{[32]} lutheran-mo synod
\item \verb{[33]} wi evan luth synod
\item \verb{[34]} other lutheran
\item \verb{[35]} evangelical luth
\item \verb{[38]} lutheran-dk which
\item \verb{[40]} presbyterian c in us
\item \verb{[41]} united pres ch in us
\item \verb{[42]} other presbyterian
\item \verb{[48]} presbyterian-dk wh
\item \verb{[50]} episcopal
\item \verb{[60]} other
\item \verb{[70]} no denomination
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1059/vshow}{GSS Data Explorer page for \code{maden}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab iap \tab am baptist asso \tab am bapt ch in usa \tab nat bapt conv of am \tab nat bapt conv usa \tab southern baptist \tab other baptists \tab baptist-dk which \tab afr meth episcopal \tab afr meth ep zion \tab united methodist \tab other methodist \tab methodist-dk which \tab am lutheran \tab luth ch in america \tab lutheran-mo synod \tab wi evan luth synod \tab other lutheran \tab lutheran-dk which \tab presbyterian c in us \tab united pres ch in us \tab other presbyterian \tab presbyterian-dk wh \tab episcopal \tab other \tab no denomination \tab no answer \tab Total \cr
   1988 \tab 566 \tab 25 \tab 9 \tab 2 \tab 6 \tab 163 \tab 35 \tab 98 \tab 9 \tab 2 \tab 116 \tab 9 \tab 43 \tab 21 \tab 12 \tab 19 \tab 3 \tab 7 \tab 33 \tab 14 \tab 18 \tab 7 \tab 30 \tab 31 \tab 161 \tab 34 \tab 8 \tab 1481 \cr
   Total \tab 566 \tab 25 \tab 9 \tab 2 \tab 6 \tab 163 \tab 35 \tab 98 \tab 9 \tab 2 \tab 116 \tab 9 \tab 43 \tab 21 \tab 12 \tab 19 \tab 3 \tab 7 \tab 33 \tab 14 \tab 18 \tab 7 \tab 30 \tab 31 \tab 161 \tab 34 \tab 8 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{marelig}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}}
}
\concept{Religion}
\keyword{variable}
