% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{madenkid}
\alias{madenkid}
\title{If protestant, mothers denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
madenkid
}
\details{
Question What was your mother's religious preference when you were a child? Was it Protestant, Catholic, Jewish, some other religion, or no religion?
A.  If Protestant, what specific denomination was that?
}
\section{Values}{

\itemize{
\item \verb{[10000]} american baptist association
\item \verb{[11000]} american baptist churches in the u.s.a.
\item \verb{[12000]} national baptist convention of america
\item \verb{[13000]} national baptist convention u.s.a., inc.
\item \verb{[14000]} southern baptist convention
\item \verb{[15000]} other baptist churches
\item \verb{[18000]} baptist, don't know which
\item \verb{[20000]} african methodist episcopal church
\item \verb{[21000]} african methodist episcopal zion church
\item \verb{[22000]} united methodist church
\item \verb{[23000]} other methodist churches
\item \verb{[28000]} methodist, don't know which
\item \verb{[30000]} american lutheran church
\item \verb{[31000]} lutheran church in america
\item \verb{[32000]} lutheran church - missouri synod
\item \verb{[33000]} wisconsin evangelical lutheran synod
\item \verb{[34000]} other lutheran churches
\item \verb{[35000]} evangelical lutheran
\item \verb{[38000]} lutheran, don't know which
\item \verb{[40000]} presbyterian church in the u.s.
\item \verb{[41000]} united presbyterian church in the u.s.a.
\item \verb{[42000]} other presbyterian churches
\item \verb{[43000]} presbyterian church u.s.a.
\item \verb{[48000]} presbyterian, don't know which
\item \verb{[50000]} episcopal church
\item \verb{[60001]} hungarian reformed
\item \verb{[60002]} evangelical congregational
\item \verb{[60003]} ind bible, bible, bible fellowship
\item \verb{[60005]} church of prophecy
\item \verb{[60006]} new testament christian
\item \verb{[60007]} church of god, saint & christ
\item \verb{[60008]} moravian
\item \verb{[60009]} christian & missionary alliance
\item \verb{[60010]} advent christian
\item \verb{[60011]} spiritualist
\item \verb{[60012]} assembly of god
\item \verb{[60013]} free methodist
\item \verb{[60014]} apostolic faith
\item \verb{[60015]} african methodist
\item \verb{[60016]} free will baptist
\item \verb{[60017]} eden evangelist
\item \verb{[60018]} holiness (nazarene)
\item \verb{[60019]} baptist (northern)
\item \verb{[60020]} brethren church, brethren
\item \verb{[60021]} witness holiness
\item \verb{[60022]} brethren, plymouth
\item \verb{[60023]} united brethren, united brethren in christ
\item \verb{[60024]} independent
\item \verb{[60025]} christian disciples
\item \verb{[60026]} christ in christian union
\item \verb{[60027]} open bible
\item \verb{[60028]} christian catholic
\item \verb{[60029]} christ church unity
\item \verb{[60030]} christ adelphians
\item \verb{[60031]} christian; central christian
\item \verb{[60032]} christian reform
\item \verb{[60033]} christian scientist
\item \verb{[60034]} church of christ, evangelical
\item \verb{[60035]} church of christ
\item \verb{[60036]} churches of god (except with christ and holiness)
\item \verb{[60037]} church of god in christ
\item \verb{[60038]} church of god in christ holiness
\item \verb{[60039]} church of the living god
\item \verb{[60040]} congregationalist, 1st congreg
\item \verb{[60041]} community church
\item \verb{[60042]} covenant
\item \verb{[60043]} dutch reform
\item \verb{[60044]} disciples of christ
\item \verb{[60045]} evangelical, evangelist
\item \verb{[60046]} evangelical reformed
\item \verb{[60047]} evangelist free church
\item \verb{[60048]} first church
\item \verb{[60049]} first christian disciples of christ
\item \verb{[60050]} first reformed
\item \verb{[60051]} first christian
\item \verb{[60052]} full gospel
\item \verb{[60053]} four square gospel
\item \verb{[60054]} friends
\item \verb{[60055]} holy roller
\item \verb{[60056]} holiness; church of holiness
\item \verb{[60057]} pilgrim holiness
\item \verb{[60058]} jehovah''s witnesses
\item \verb{[60059]} lds
\item \verb{[60060]} lds--mormon
\item \verb{[60061]} lds--reorganized
\item \verb{[60062]} lds--jesus christ; church of jesus lds
\item \verb{[60063]} mennonite
\item \verb{[60064]} mormon
\item \verb{[60065]} nazarene
\item \verb{[60066]} pentecostal assembly of god
\item \verb{[60067]} pentecostal church of god
\item \verb{[60068]} pentecostal
\item \verb{[60069]} pentecostal holiness, holiness pentecostal
\item \verb{[60070]} quaker
\item \verb{[60071]} reformed
\item \verb{[60072]} reformed united church of christ
\item \verb{[60073]} reformed church of christ
\item \verb{[60074]} religious science
\item \verb{[60075]} mind science
\item \verb{[60076]} salvation army
\item \verb{[60077]} 7th day adventist
\item \verb{[60078]} sanctified, sanctification
\item \verb{[60079]} united holiness
\item \verb{[60080]} unitarian, universalist
\item \verb{[60081]} united church of christ
\item \verb{[60082]} united church, unity church
\item \verb{[60083]} wesleyan
\item \verb{[60084]} wesleyan methodist--pilgrim
\item \verb{[60085]} zion union
\item \verb{[60086]} zion union apostolic
\item \verb{[60087]} zion union apostolic--reformed
\item \verb{[60088]} disciples of god
\item \verb{[60089]} grace reformed
\item \verb{[60090]} holiness church of god
\item \verb{[60091]} evangelical covenant
\item \verb{[60092]} mission covenant
\item \verb{[60093]} missionary baptist
\item \verb{[60094]} swedish mission
\item \verb{[60095]} unity
\item \verb{[60096]} united church of christianity
\item \verb{[60097]} other fundamentalist
\item \verb{[60098]} federated church
\item \verb{[60099]} american reform
\item \verb{[60100]} grace brethren
\item \verb{[60101]} christ in god
\item \verb{[60102]} charismatic
\item \verb{[60103]} pentecostal apostolic
\item \verb{[60104]} house of prayer
\item \verb{[60105]} latvian lutheran
\item \verb{[60106]} triumph church of god
\item \verb{[60107]} apostolic christian
\item \verb{[60108]} christ cathedral of truth
\item \verb{[60109]} bible missionary
\item \verb{[60110]} calvary bible
\item \verb{[60111]} amish
\item \verb{[60112]} evangelical methodist
\item \verb{[60113]} worldwide church of god
\item \verb{[60114]} church universal and triumphant
\item \verb{[60115]} mennonite brethren
\item \verb{[60116]} church of the first born
\item \verb{[60117]} missionary church
\item \verb{[60118]} the way ministry
\item \verb{[60119]} united church of canada
\item \verb{[60120]} evangelical united brethren
\item \verb{[60121]} the church of god of prophecy
\item \verb{[60122]} chapel of faith
\item \verb{[60123]} polish national church
\item \verb{[60124]} faith gospel tabernacle
\item \verb{[60125]} christian calvary chapel
\item \verb{[60126]} carmelite
\item \verb{[60127]} church of daniel's band
\item \verb{[60128]} christian tabernacle
\item \verb{[60129]} living word
\item \verb{[60130]} true light church of christ
\item \verb{[60131]} macedonia
\item \verb{[60132]} brother of christ
\item \verb{[60133]} primitive baptist
\item \verb{[60134]} independent fundamental church of america
\item \verb{[60135]} chinese gospel church
\item \verb{[60136]} new age spirituality
\item \verb{[60137]} new song
\item \verb{[60138]} apostolic church
\item \verb{[60139]} faith christian
\item \verb{[60140]} people's church
\item \verb{[60141]} new birth christian
\item \verb{[60142]} unity school of christianity
\item \verb{[60143]} assyrian evangelist church
\item \verb{[60144]} spirit of christ
\item \verb{[60145]} church of jesus christ of the restoration
\item \verb{[60146]} laotian christian
\item \verb{[60148]} schwenkfelder
\item \verb{[60149]} polish catholic
\item \verb{[60150]} zwinglian
\item \verb{[60151]} world overcomer outreach ministry
\item \verb{[60152]} course in miracles
\item \verb{[60153]} unity of the brethren
\item \verb{[60154]} spirit filled
\item \verb{[60155]} christian union
\item \verb{[60156]} church of living christ
\item \verb{[60157]} community of christ
\item \verb{[60158]} new hope christian fellowship
\item \verb{[60159]} community christian fellowship
\item \verb{[60160]} friends in christ
\item \verb{[60161]} hawaiian ohana
\item \verb{[60162]} reformed church of jesus christ of latter day saints
\item \verb{[60163]} swedenurgian/churches of the new jerusalem
\item \verb{[60164]} divine science
\item \verb{[60166]} united christian
\item \verb{[60167]} sanctuary
\item \verb{[60168]} rain on us deliverance ministries
\item \verb{[60169]} the word church
\item \verb{[60170]} cornerstone church
\item \verb{[60171]} life sanctuary
\item \verb{[60172]} word of faith church
\item \verb{[60173]} harvest church
\item \verb{[60174]} shephard''s chapel
\item \verb{[60175]} greater new testament church
\item \verb{[60176]} vineyard church
\item \verb{[60177]} real life ministries
\item \verb{[60178]} cathedral of joy
\item \verb{[60179]} great faith ministries
\item \verb{[60180]} shield of faith ministries
\item \verb{[60181]} born again
\item \verb{[60182]} alliance
\item \verb{[60183]} jacobite apostolic
\item \verb{[60184]} church of god of israel
\item \verb{[60185]} journeys
\item \verb{[60186]} national progressive baptist
\item \verb{[60187]} new apostolic
\item \verb{[60188]} metropolitan community
\item \verb{[60189]} family life church
\item \verb{[60190]} faith fellowship
\item \verb{[60191]} faith covenant
\item \verb{[60192]} new thought
\item \verb{[60193]} free spirit ministry
\item \verb{[60194]} jacobite syrian christian church
\item \verb{[60195]} the ark church
\item \verb{[60196]} empowerment temple
\item \verb{[60197]} grace independent baptist church
\item \verb{[60198]} new life
\item \verb{[60199]} follower of christ
\item \verb{[60200]} north country chapel
\item \verb{[60201]} pathways christian church
\item \verb{[60202]} messianic christian
\item \verb{[60203]} warehouse ministries
\item \verb{[60204]} universal church of the kingdom of god
\item \verb{[60205]} assembly of christ
\item \verb{[60206]} the amana church
\item \verb{[60207]} the legacy church
\item \verb{[60208]} calvary
\item \verb{[60209]} church of the living god
\item \verb{[60210]} ethiopian evangelical church
\item \verb{[60211]} disciple of jesus
\item \verb{[60212]} scandinavian church
\item \verb{[60213]} hebrew roots
\item \verb{[60214]} hebrew isrealites
\item \verb{[60215]} armenian apostolic church
\item \verb{[60216]} word of jesus
\item \verb{[60217]} old testament
\item \verb{[70000]} no denomination given or non-denom. church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4666/vshow}{GSS Data Explorer page for \code{madenkid}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab american baptist association \tab southern baptist convention \tab other baptist churches \tab baptist, don't know which \tab african methodist episcopal church \tab african methodist episcopal zion church \tab united methodist church \tab other methodist churches \tab methodist, don't know which \tab american lutheran church \tab lutheran church - missouri synod \tab wisconsin evangelical lutheran synod \tab other lutheran churches \tab lutheran, don't know which \tab united presbyterian church in the u.s.a. \tab presbyterian church u.s.a. \tab presbyterian, don't know which \tab episcopal church \tab new testament christian \tab moravian \tab assembly of god \tab free will baptist \tab christian; central christian \tab christian reform \tab christian scientist \tab church of christ \tab churches of god (except with christ and holiness) \tab church of god in christ \tab congregationalist, 1st congreg \tab covenant \tab dutch reform \tab disciples of christ \tab evangelical, evangelist \tab first christian disciples of christ \tab first reformed \tab first christian \tab full gospel \tab holiness; church of holiness \tab pilgrim holiness \tab jehovah''s witnesses \tab lds \tab lds--mormon \tab lds--reorganized \tab lds--jesus christ; church of jesus lds \tab mennonite \tab mormon \tab nazarene \tab pentecostal assembly of god \tab pentecostal \tab quaker \tab reformed \tab 7th day adventist \tab sanctified, sanctification \tab unitarian, universalist \tab united church of christ \tab wesleyan \tab wesleyan methodist--pilgrim \tab mission covenant \tab missionary baptist \tab other fundamentalist \tab worldwide church of god \tab united church of canada \tab evangelical united brethren \tab primitive baptist \tab new birth christian \tab no denomination given or non-denom. church \tab don't know \tab no answer \tab lutheran church in america \tab evangelical lutheran \tab independent \tab church of god in christ holiness \tab evangelical reformed \tab salvation army \tab apostolic church \tab american baptist churches in the u.s.a. \tab national baptist convention of america \tab national baptist convention u.s.a., inc. \tab presbyterian church in the u.s. \tab ind bible, bible, bible fellowship \tab christian & missionary alliance \tab apostolic faith \tab brethren church, brethren \tab united brethren, united brethren in christ \tab charismatic \tab missionary church \tab armenian apostolic church \tab not available in this year \tab Total \cr
   1991 \tab 707 \tab 2 \tab 34 \tab 3 \tab 182 \tab 3 \tab 1 \tab 19 \tab 1 \tab 99 \tab 6 \tab 4 \tab 1 \tab 2 \tab 72 \tab 2 \tab 2 \tab 44 \tab 23 \tab 1 \tab 1 \tab 7 \tab 2 \tab 18 \tab 2 \tab 4 \tab 14 \tab 10 \tab 2 \tab 8 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab 3 \tab 15 \tab 1 \tab 1 \tab 2 \tab 3 \tab 4 \tab 2 \tab 1 \tab 17 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 12 \tab 13 \tab 120 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1998 \tab 2126 \tab 1 \tab 42 \tab 5 \tab 159 \tab 27 \tab 1 \tab 23 \tab 3 \tab 93 \tab 1 \tab 3 \tab 2 \tab 2 \tab 49 \tab - \tab - \tab 38 \tab 10 \tab - \tab 1 \tab 5 \tab - \tab 16 \tab 1 \tab 7 \tab 9 \tab 6 \tab 2 \tab 8 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 10 \tab - \tab - \tab - \tab 1 \tab - \tab 8 \tab 2 \tab - \tab 8 \tab - \tab 1 \tab 12 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 16 \tab 20 \tab 83 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2008 \tab 1254 \tab 15 \tab 103 \tab 2 \tab 135 \tab 5 \tab 3 \tab 66 \tab - \tab 30 \tab 10 \tab 11 \tab 5 \tab 1 \tab 30 \tab 7 \tab - \tab 37 \tab 30 \tab - \tab - \tab 3 \tab 6 \tab 27 \tab 3 \tab - \tab 15 \tab 6 \tab 1 \tab 5 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 9 \tab 8 \tab 4 \tab - \tab - \tab 2 \tab 11 \tab 3 \tab - \tab 19 \tab 3 \tab 4 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 41 \tab 31 \tab 13 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 6 \tab 3 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2023 \cr
   2018 \tab 1748 \tab 2 \tab 15 \tab 1 \tab 186 \tab 1 \tab 1 \tab 12 \tab 1 \tab 58 \tab - \tab 4 \tab 1 \tab - \tab 42 \tab - \tab 1 \tab 27 \tab 21 \tab 1 \tab - \tab 3 \tab 4 \tab 55 \tab - \tab - \tab 13 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 12 \tab - \tab - \tab - \tab - \tab 3 \tab 14 \tab 1 \tab - \tab 21 \tab 3 \tab 2 \tab 7 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 24 \tab 27 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 2348 \cr
   Total \tab 5835 \tab 20 \tab 194 \tab 11 \tab 662 \tab 36 \tab 6 \tab 120 \tab 5 \tab 280 \tab 17 \tab 22 \tab 9 \tab 5 \tab 193 \tab 9 \tab 3 \tab 146 \tab 84 \tab 2 \tab 2 \tab 18 \tab 12 \tab 116 \tab 6 \tab 11 \tab 51 \tab 25 \tab 5 \tab 25 \tab 3 \tab 2 \tab 11 \tab 11 \tab 1 \tab 1 \tab 5 \tab 5 \tab 14 \tab 2 \tab 34 \tab 23 \tab 5 \tab 1 \tab 3 \tab 8 \tab 37 \tab 8 \tab 1 \tab 65 \tab 8 \tab 8 \tab 26 \tab 1 \tab 5 \tab 11 \tab 3 \tab 2 \tab 1 \tab 10 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2 \tab 1 \tab 93 \tab 91 \tab 219 \tab 8 \tab 5 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 5 \tab 6 \tab 3 \tab 6 \tab 1 \tab 3 \tab 1 \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab 0 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{befair}},
\code{\link{bible1}},
\code{\link{cantrust}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conbiz}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{hapunhap}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relcand}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}

Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numbwfrd}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{othsch}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcand}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
