% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{maeduc}
\alias{maeduc}
\title{Highest year school completed, mother (or youngest same-sex parent)}
\source{
General Social Survey https://gss.norc.org
}
\description{
maeduc
}
\details{
Question RESPONDENT'S MOTHER'S (SUBSTITUTE MOTHER'S) EDUCATION
}
\section{Values}{

\itemize{
\item \code{0} no formal schooling
\item \code{1} 1st grade
\item \code{2} 2nd grade
\item \code{3} 3rd grade
\item \code{4} 4th grade
\item \code{5} 5th grade
\item \code{6} 6th grade
\item \code{7} 7th grade
\item \code{8} 8th grade
\item \code{9} 9th grade
\item \code{10} 10th grade
\item \code{11} 11th grade
\item \code{12} 12th grade
\item \code{13} 1 year of college
\item \code{14} 2 years of college
\item \code{15} 3 years of college
\item \code{16} 4 years of college
\item \code{17} 5 years of college
\item \code{18} 6 years of college
\item \code{19} 7 years of college
\item \code{20} 8 or more years of college
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/57/vshow}{GSS Data Explorer page for \code{maeduc}}.
}

\seealso{
Other Core: 
\code{\link{age}},
\code{\link{agekdbrn}},
\code{\link{codeg}},
\code{\link{coeduc}},
\code{\link{degree}},
\code{\link{educ}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{speduc}}
}
\concept{Core}
\keyword{variable}
