% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{mafund}
\alias{mafund}
\title{How fundamentalist was r's mother}
\source{
General Social Survey https://gss.norc.org
}
\description{
mafund
}
\details{
Question C. Fundamentalism/Liberalism of Mother's Religion
}
\section{Values}{

\itemize{
\item \verb{[1]} fundamentalist
\item \verb{[2]} moderate
\item \verb{[3]} liberal
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1062/vshow}{GSS Data Explorer page for \code{mafund}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab fundamentalist \tab moderate \tab liberal \tab no answer \tab Total \cr
   1988 \tab 48 \tab 475 \tab 566 \tab 343 \tab 49 \tab 1481 \cr
   Total \tab 48 \tab 475 \tab 566 \tab 343 \tab 49 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
