% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{maind10}
\alias{maind10}
\title{Mothers industry code (naics 2007)}
\source{
General Social Survey https://gss.norc.org
}
\description{
maind10
}
\details{
Question A. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up?  That is, what was her job called?
OCCUPATION: (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What did she actually do in that job? Tell me, what were some of her main duties?
C. What kind of place did she work for?
INDUSTRY: (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was she self-employed, or did she work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[170]} crop production
\item \verb{[180]} animal production
\item \verb{[190]} forestry, except logging
\item \verb{[270]} logging
\item \verb{[280]} fishing, hunting, and trapping
\item \verb{[290]} support activities for agriculture and forestry
\item \verb{[370]} oil and gas extraction
\item \verb{[380]} coal mining
\item \verb{[390]} metal ore mining
\item \verb{[470]} nonmetallic mineral mining and quarrying
\item \verb{[480]} not specified type of mining
\item \verb{[490]} support activities for mining
\item \verb{[570]} electric power generation, transmission, and distribution
\item \verb{[580]} natural gas distribution
\item \verb{[590]} electric and gas, and other combinations
\item \verb{[670]} water, steam, air-conditioning, and irrigation systems
\item \verb{[680]} sewage treatment facilities
\item \verb{[690]} not specified utilities
\item \verb{[770]} construction
\item \verb{[1070]} animal food, grain, and oilseed milling
\item \verb{[1080]} sugar and confectionery products
\item \verb{[1090]} fruit and vegetable preserving and specialty food manufacturing
\item \verb{[1170]} dairy product manufacturing
\item \verb{[1180]} animal slaughtering and processing
\item \verb{[1190]} retail bakeries
\item \verb{[1270]} bakeries, except retail
\item \verb{[1280]} seafood and other miscellaneous foods, n.e.c.
\item \verb{[1290]} not specified food industries
\item \verb{[1370]} beverage manufacturing
\item \verb{[1390]} tobacco manufacturing
\item \verb{[1470]} fiber, yarn, and thread mills
\item \verb{[1480]} fabric mills, except knitting mills
\item \verb{[1490]} textile and fabric finishing and coating mills
\item \verb{[1570]} carpet and rug mills
\item \verb{[1590]} textile product mills, except carpet and rug
\item \verb{[1670]} knitting fabric mills, and apparel knitting mills
\item \verb{[1680]} cut and sew apparel manufacturing
\item \verb{[1690]} apparel accessories and other apparel manufacturing
\item \verb{[1770]} footwear manufacturing
\item \verb{[1790]} leather tanning and finishing and other allied products manufacturing
\item \verb{[1870]} pulp, paper, and paperboard mills
\item \verb{[1880]} paperboard containers and boxes
\item \verb{[1890]} miscellaneous paper and pulp products
\item \verb{[1990]} printing and related support activities
\item \verb{[2070]} petroleum refining
\item \verb{[2090]} miscellaneous petroleum and coal products
\item \verb{[2170]} resin, synthetic rubber and fibers, and filaments manufacturing
\item \verb{[2180]} agricultural chemical manufacturing
\item \verb{[2190]} pharmaceutical and medicine manufacturing
\item \verb{[2270]} paint, coating, and adhesive manufacturing
\item \verb{[2280]} soap, cleaning compound, and cosmetics manufacturing
\item \verb{[2290]} industrial and miscellaneous chemicals
\item \verb{[2370]} plastics product manufacturing
\item \verb{[2380]} tire manufacturing
\item \verb{[2390]} rubber product, except tire, manufacturing
\item \verb{[2470]} pottery, ceramics, and plumbing fixture manufacturing
\item \verb{[2480]} structural clay product manufacturing
\item \verb{[2490]} glass and glass product manufacturing
\item \verb{[2570]} cement, concrete, lime, and gypsum product manufacturing
\item \verb{[2590]} miscellaneous nonmetallic mineral product manufacturing
\item \verb{[2670]} iron and steel mills and steel product manufacturing
\item \verb{[2680]} aluminum production and processing
\item \verb{[2690]} nonferrous metal (except aluminum) production and processing
\item \verb{[2770]} foundries
\item \verb{[2780]} metal forgings and stampings
\item \verb{[2790]} cutlery and hand tool manufacturing
\item \verb{[2870]} structural metals, and boiler, tank, and shipping container manufacturing
\item \verb{[2880]} machine shops; turned product; screw, nut, and bolt manufacturing
\item \verb{[2890]} coating, engraving, heat treating and allied activities
\item \verb{[2970]} ordnance
\item \verb{[2980]} miscellaneous fabricated metal products manufacturing
\item \verb{[2990]} not specified metal industries
\item \verb{[3070]} agricultural implement manufacturing
\item \verb{[3080]} construction, and mining and oil and gas field machinery manufacturing
\item \verb{[3090]} commercial and service industry machinery manufacturing
\item \verb{[3170]} metalworking machinery manufacturing
\item \verb{[3180]} engines, turbines, and power transmission equipment manufacturing
\item \verb{[3190]} machinery manufacturing, n.e.c.
\item \verb{[3290]} not specified machinery manufacturing
\item \verb{[3360]} computer and peripheral equipment manufacturing
\item \verb{[3370]} communications, and audio and video equipment manufacturing
\item \verb{[3380]} navigational, measuring, electromedical, and control instruments manufacturing
\item \verb{[3390]} electronic component and product manufacturing, n.e.c.
\item \verb{[3470]} household appliance manufacturing
\item \verb{[3490]} electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.
\item \verb{[3570]} motor vehicles and motor vehicle equipment manufacturing
\item \verb{[3580]} aircraft and parts manufacturing
\item \verb{[3590]} aerospace product and parts manufacturing
\item \verb{[3670]} railroad rolling stock manufacturing
\item \verb{[3680]} ship and boat building
\item \verb{[3690]} other transportation equipment manufacturing
\item \verb{[3770]} sawmills and wood preservation
\item \verb{[3780]} veneer, plywood, and engineered wood products
\item \verb{[3790]} prefabricated wood buildings and mobile homes
\item \verb{[3870]} miscellaneous wood products
\item \verb{[3890]} furniture and related product manufacturing
\item \verb{[3960]} medical equipment and supplies manufacturing
\item \verb{[3970]} sporting and athletic goods, and doll, toy and game manufacturing
\item \verb{[3980]} miscellaneous manufacturing, n.e.c.
\item \verb{[3990]} not specified manufacturing industries
\item \verb{[4070]} motor vehicles, parts and supplies, merchant wholesalers
\item \verb{[4080]} furniture and home furnishing, merchant wholesalers
\item \verb{[4090]} lumber and other construction materials, merchant wholesalers
\item \verb{[4170]} professional and commercial equipment and supplies, merchant wholesalers
\item \verb{[4180]} metals and minerals, except petroleum, merchant wholesalers
\item \verb{[4190]} electrical and electronic goods, merchant wholesalers
\item \verb{[4260]} hardware, plumbing and heating equipment, and supplies, merchant wholesalers
\item \verb{[4270]} machinery, equipment, and supplies, merchant wholesalers
\item \verb{[4280]} recyclable material, merchant wholesalers
\item \verb{[4290]} miscellaneous durable goods, merchant wholesalers
\item \verb{[4370]} paper and paper products, merchant wholesalers
\item \verb{[4380]} drugs, sundries, and chemical and allied products, merchant wholesalers
\item \verb{[4390]} apparel, fabrics, and notions, merchant wholesalers
\item \verb{[4470]} groceries and related products, merchant wholesalers
\item \verb{[4480]} farm product raw materials, merchant wholesalers
\item \verb{[4490]} petroleum and petroleum products, merchant wholesalers
\item \verb{[4560]} alcoholic beverages, merchant wholesalers
\item \verb{[4570]} farm supplies, merchant wholesalers
\item \verb{[4580]} miscellaneous nondurable goods, merchant wholesalers
\item \verb{[4585]} wholesale electronic markets, agents and brokers
\item \verb{[4590]} not specified wholesale trade
\item \verb{[4670]} automobile dealers
\item \verb{[4680]} other motor vehicle dealers
\item \verb{[4690]} auto parts, accessories, and tire stores
\item \verb{[4770]} furniture and home furnishings stores
\item \verb{[4780]} household appliance stores
\item \verb{[4790]} radio, tv, and computer stores
\item \verb{[4870]} building material and supplies dealers
\item \verb{[4880]} hardware stores
\item \verb{[4890]} lawn and garden equipment and supplies stores
\item \verb{[4970]} grocery stores
\item \verb{[4980]} specialty food stores
\item \verb{[4990]} beer, wine, and liquor stores
\item \verb{[5070]} pharmacies and drug stores
\item \verb{[5080]} health and personal care, except drug, stores
\item \verb{[5090]} gasoline stations
\item \verb{[5170]} clothing stores
\item \verb{[5180]} shoe stores
\item \verb{[5190]} jewelry, luggage, and leather goods stores
\item \verb{[5270]} sporting goods, camera, and hobby and toy stores
\item \verb{[5280]} sewing, needlework, and piece goods stores
\item \verb{[5290]} music stores
\item \verb{[5370]} book stores and news dealers
\item \verb{[5380]} department stores and discount stores
\item \verb{[5390]} miscellaneous general merchandise stores
\item \verb{[5470]} retail florists
\item \verb{[5480]} office supplies and stationery stores
\item \verb{[5490]} used merchandise stores
\item \verb{[5570]} gift, novelty, and souvenir shops
\item \verb{[5580]} miscellaneous retail stores
\item \verb{[5590]} electronic shopping
\item \verb{[5591]} electronic auctions
\item \verb{[5592]} mail order houses
\item \verb{[5670]} vending machine operators
\item \verb{[5680]} fuel dealers
\item \verb{[5690]} other direct selling establishments
\item \verb{[5790]} not specified retail trade
\item \verb{[6070]} air transportation
\item \verb{[6080]} rail transportation
\item \verb{[6090]} water transportation
\item \verb{[6170]} truck transportation
\item \verb{[6180]} bus service and urban transit
\item \verb{[6190]} taxi and limousine service
\item \verb{[6270]} pipeline transportation
\item \verb{[6280]} scenic and sightseeing transportation
\item \verb{[6290]} services incidental to transportation
\item \verb{[6370]} postal service
\item \verb{[6380]} couriers and messengers
\item \verb{[6390]} warehousing and storage
\item \verb{[6470]} newspaper publishers
\item \verb{[6480]} periodical, book, and directory publishers
\item \verb{[6490]} software publishers
\item \verb{[6570]} motion pictures and video industries
\item \verb{[6590]} sound recording industries
\item \verb{[6670]} radio and television broadcasting and cable subscription programming
\item \verb{[6672]} internet publishing and broadcasting and web search portals
\item \verb{[6680]} wired telecommunications carriers
\item \verb{[6690]} other telecommunications services
\item \verb{[6695]} data processing, hosting, and related services
\item \verb{[6770]} libraries and archives
\item \verb{[6780]} other information services
\item \verb{[6870]} banking and related activities
\item \verb{[6880]} savings institutions, including credit unions
\item \verb{[6890]} non-depository credit and related activities
\item \verb{[6970]} securities, commodities, funds, trusts, and other financial investments
\item \verb{[6990]} insurance carriers and related activities
\item \verb{[7070]} real estate
\item \verb{[7080]} automotive equipment rental and leasing
\item \verb{[7170]} video tape and disk rental
\item \verb{[7180]} other consumer goods rental
\item \verb{[7190]} commercial, industrial, and other intangible assets rental and leasing
\item \verb{[7270]} legal services
\item \verb{[7280]} accounting, tax preparation, bookkeeping, and payroll services
\item \verb{[7290]} architectural, engineering, and related services
\item \verb{[7370]} specialized design services
\item \verb{[7380]} computer systems design and related services
\item \verb{[7390]} management, scientific, and technical consulting services
\item \verb{[7460]} scientific research and development services
\item \verb{[7470]} advertising and related services
\item \verb{[7480]} veterinary services
\item \verb{[7490]} other professional, scientific, and technical services
\item \verb{[7570]} management of companies and enterprises
\item \verb{[7580]} employment services
\item \verb{[7590]} business support services
\item \verb{[7670]} travel arrangements and reservation services
\item \verb{[7680]} investigation and security services
\item \verb{[7690]} services to buildings and dwellings
\item \verb{[7770]} landscaping services
\item \verb{[7780]} other administrative and other support services
\item \verb{[7790]} waste management and remediation services
\item \verb{[7860]} elementary and secondary schools
\item \verb{[7870]} colleges and universities, including junior colleges
\item \verb{[7880]} business, technical, and trade schools and training
\item \verb{[7890]} other schools and instruction, and educational support services
\item \verb{[7970]} offices of physicians
\item \verb{[7980]} offices of dentists
\item \verb{[7990]} offices of chiropractors
\item \verb{[8070]} offices of optometrists
\item \verb{[8080]} offices of other health practitioners
\item \verb{[8090]} outpatient care centers
\item \verb{[8170]} home health care services
\item \verb{[8180]} other health care services
\item \verb{[8190]} hospitals
\item \verb{[8270]} nursing care facilities
\item \verb{[8290]} residential care facilities, without nursing
\item \verb{[8370]} individual and family services
\item \verb{[8380]} community food and housing, and emergency services
\item \verb{[8390]} vocational rehabilitation services
\item \verb{[8470]} child day care services
\item \verb{[8560]} independent artists, performing arts, spectator sports, and related industries
\item \verb{[8570]} museums, art galleries, historical sites, and similar institutions
\item \verb{[8580]} bowling centers
\item \verb{[8590]} other amusement, gambling, and recreation industries
\item \verb{[8660]} traveler accommodation
\item \verb{[8670]} recreational vehicle parks and camps, and rooming and boarding houses
\item \verb{[8680]} restaurants and other food services
\item \verb{[8690]} drinking places, alcoholic beverages
\item \verb{[8770]} automotive repair and maintenance
\item \verb{[8780]} car washes
\item \verb{[8790]} electronic and precision equipment repair and maintenance
\item \verb{[8870]} commercial and industrial machinery and equipment repair and maintenance
\item \verb{[8880]} personal and household goods repair and maintenance
\item \verb{[8890]} footwear and leather goods repair
\item \verb{[8970]} barber shops
\item \verb{[8980]} beauty salons
\item \verb{[8990]} nail salons and other personal care services
\item \verb{[9070]} drycleaning and laundry services
\item \verb{[9080]} funeral homes, and cemeteries and crematories
\item \verb{[9090]} other personal services
\item \verb{[9160]} religious organizations
\item \verb{[9170]} civic, social, advocacy organizations, and grantmaking and giving services
\item \verb{[9180]} labor unions
\item \verb{[9190]} business, professional, political, and similar organizations
\item \verb{[9290]} private households
\item \verb{[9370]} executive offices and legislative bodies
\item \verb{[9380]} public finance activities
\item \verb{[9390]} other general government and support
\item \verb{[9470]} justice, public order, and safety activities
\item \verb{[9480]} administration of human resource programs
\item \verb{[9490]} administration of environmental quality and housing programs
\item \verb{[9570]} administration of economic programs and space research
\item \verb{[9590]} national security and international affairs
\item \verb{[9670]} u.s. army
\item \verb{[9680]} u.s. air force
\item \verb{[9690]} u.s. navy
\item \verb{[9770]} u.s. marine corps
\item \verb{[9780]} u.s. coast guard
\item \verb{[9790]} u.s. armed forces/n.s.
\item \verb{[9870]} active duty military personnel, all other
\item \verb{[9890]} uncodeable
\item \verb{[9997]} uncodeable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abhave3}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{ablegal}},
\code{\link{abnomore}},
\code{\link{aborct}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{abspno}},
\code{\link{adoptus}},
\code{\link{adults}},
\code{\link{affrmact}},
\code{\link{age}},
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{agewed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{antiwar}},
\code{\link{arrest}},
\code{\link{attend}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{barate}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blksimp}},
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{bored}},
\code{\link{born}},
\code{\link{brass}},
\code{\link{broadband}},
\code{\link{burglr}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{capfirm}},
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{cathtemp}},
\code{\link{childs}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chlddth}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldsoon}},
\code{\link{cigweek}},
\code{\link{civic}},
\code{\link{civright}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{cobarate}},
\code{\link{codeg}},
\code{\link{coden}},
\code{\link{coden_1222}},
\code{\link{codipged}},
\code{\link{coeduc}},
\code{\link{coeftotlt}},
\code{\link{coevwork}},
\code{\link{cofund}},
\code{\link{cogradtounder}},
\code{\link{cohrs1}},
\code{\link{cohrs2}},
\code{\link{coind10}},
\code{\link{cojail}},
\code{\link{cojew}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{color}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{commute}},
\code{\link{comprend}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlabory}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conorelgsp}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{contemp}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{convictd}},
\code{\link{coocc10}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{coother}},
\code{\link{copres10}},
\code{\link{copunish}},
\code{\link{corel}},
\code{\link{cosector}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{defwrkev}},
\code{\link{defwrknw}},
\code{\link{degree}},
\code{\link{dejavu}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{dotdata}},
\code{\link{dotged}},
\code{\link{dotpeop}},
\code{\link{dotpres}},
\code{\link{dotsvp}},
\code{\link{dotthng}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{earnrs}},
\code{\link{educ}},
\code{\link{educop}},
\code{\link{eftotlt}},
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{era}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{esp}},
\code{\link{evsmoke}},
\code{\link{evwork}},
\code{\link{fair}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{family16sex}},
\code{\link{father}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{febear}},
\code{\link{febrass}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fedirty}},
\code{\link{fefam}},
\code{\link{fefight}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{fehelp}},
\code{\link{fehlpmil}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{femilop}},
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{fenumok}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrite}},
\code{\link{fightair}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{form}},
\code{\link{found}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govcare}},
\code{\link{grace}},
\code{\link{gradtounder}},
\code{\link{granborn}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{guerilla}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}},
\code{\link{gunner}},
\code{\link{gunnum}},
\code{\link{hapcohab}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{healer}},
\code{\link{health}},
\code{\link{helpblk}},
\code{\link{helpful}},
\code{\link{helpnot}},
\code{\link{helpoth}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitnum}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{id}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{immcrime}},
\code{\link{immjobs}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imppol}},
\code{\link{imprelax}},
\code{\link{impwork}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus07}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewtemp}},
\code{\link{jobaccmp}},
\code{\link{jobfind}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{jobtrain}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kidssol}},
\code{\link{king}},
\code{\link{landwar}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{lessnuke}},
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{liberatr}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{libtemp}},
\code{\link{life}},
\code{\link{likediff}},
\code{\link{lockedup}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madeath}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{maind80}},
\code{\link{maindstatus}},
\code{\link{maindtag}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{maocctag}},
\code{\link{mapa}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{mapres80}},
\code{\link{marital}},
\code{\link{marsame}},
\code{\link{marsame1}},
\code{\link{marsamey}},
\code{\link{martype}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{mawrkslf}},
\code{\link{mechanic}},
\code{\link{meserve}},
\code{\link{migage}},
\code{\link{mighist}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{mindbody}},
\code{\link{mininc}},
\code{\link{minmilop}},
\code{\link{mobile16}},
\code{\link{mobiledata}},
\code{\link{morenuke}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{neargod}},
\code{\link{news}()},
\code{\link{nonuke}},
\code{\link{norelgsp16}},
\code{\link{nukewar}},
\code{\link{nurse}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{oth16}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{owngun}},
\code{\link{paattend}},
\code{\link{padeath}},
\code{\link{padeg}},
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{paeduc}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{palefull}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres10}},
\code{\link{papres105plus}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{parborn}},
\code{\link{parsol}},
\code{\link{partfull}},
\code{\link{partyid}},
\code{\link{pawrkgrw}},
\code{\link{pawrkslf}},
\code{\link{perla}},
\code{\link{phone}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{pistol}},
\code{\link{polabuse}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{polhitok}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{polmurdr}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{popespks}},
\code{\link{popular}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{premarsx}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prestg10}},
\code{\link{prestg105plus}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{privacy}},
\code{\link{prottemp}},
\code{\link{prowar}},
\code{\link{quitsmk}},
\code{\link{racavoid}},
\code{\link{raccare}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdif5}},
\code{\link{racdif5x}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs15}},
\code{\link{raceacs2}},
\code{\link{raceacs4}},
\code{\link{raceacs6}},
\code{\link{raceacs8}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{racjoin}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{racthink}},
\code{\link{ractrust}},
\code{\link{racwrite}},
\code{\link{radiohrs}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{reg16}},
\code{\link{reg16_7222}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{relgrade}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{res16}},
\code{\link{resdefwk}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{restact}},
\code{\link{richwork}},
\code{\link{rifle}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{robbry}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{rowngun}},
\code{\link{rushed}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfin}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{satjob}},
\code{\link{school}},
\code{\link{sector}},
\code{\link{sex}},
\code{\link{sexeduc}},
\code{\link{sexnow2}},
\code{\link{shotgun}},
\code{\link{sibdeath}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{size}},
\code{\link{smoke}},
\code{\link{smokecig}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{spanking}},
\code{\link{spattend}},
\code{\link{spbarate}},
\code{\link{spdeath}},
\code{\link{spdeg}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spdipged}},
\code{\link{spdotdat}},
\code{\link{spdotged}},
\code{\link{spdotpeo}},
\code{\link{spdotpre}},
\code{\link{spdotsvp}},
\code{\link{spdotthn}},
\code{\link{spdue}},
\code{\link{speduc}},
\code{\link{speftotlt}},
\code{\link{spevwork}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spgradtounder}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spirits}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{spnorelgsp}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sppaid}},
\code{\link{sppres}},
\code{\link{sppres10}},
\code{\link{sppres105plus}},
\code{\link{sppres80}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{spsector}},
\code{\link{spwrkslf}},
\code{\link{spwrksta}},
\code{\link{srcbelt}},
\code{\link{stresswk}},
\code{\link{strike}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{teenpill}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{teensex}},
\code{\link{thnkself}},
\code{\link{ticket}},
\code{\link{tithing}},
\code{\link{transair}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{tvhours}},
\code{\link{typist}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{upgrade}()},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{vetaid}},
\code{\link{vetfam}},
\code{\link{vetfamnw}},
\code{\link{visions}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{weekswrk}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{widowed}},
\code{\link{wirtap}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}},
\code{\link{workhard}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{wrkwayup}},
\code{\link{xmarsex}},
\code{\link{xmovie}},
\code{\link{xmovie1}},
\code{\link{xnorcsiz}},
\code{\link{year}()},
\code{\link{youngen}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{Core}
\concept{Occupation}
\keyword{variable}
