% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{maind80}
\alias{maind80}
\title{Mothers industry code (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
maind80
}
\details{
Question A. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up?  That is, what was her job called?
OCCUPATION: (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What did she actually do in that job? Tell me, what were some of her main duties?
C. What kind of place did she work for?
INDUSTRY: (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was she self-employed, or did she work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[10]} agricultural production, crops
\item \verb{[11]} agricultural production, livestock
\item \verb{[20]} agricultural services, except horticultural
\item \verb{[21]} horticultural services
\item \verb{[30]} forestry
\item \verb{[31]} fishing, hunting, and trapping, etc.
\item \verb{[40]} metal mining
\item \verb{[41]} coal mining
\item \verb{[42]} crude petroleum and natural gas extraction
\item \verb{[50]} nonmetallic mining and quarrying, except fuel
\item \verb{[60]} construction
\item \verb{[100]} meat products
\item \verb{[101]} dairy products
\item \verb{[102]} canned and preserved fruits and vegetables
\item \verb{[110]} grain mill products
\item \verb{[111]} bakery products
\item \verb{[112]} sugar and confectionery products
\item \verb{[120]} beverage industries
\item \verb{[121]} miscellaneous food preparations and kindred products
\item \verb{[122]} not specified food industries
\item \verb{[130]} tobacco manufactures
\item \verb{[132]} knitting mills
\item \verb{[140]} dyeing and finishing textiles, except wool and knit goods
\item \verb{[141]} floor coverings, except hard surface
\item \verb{[142]} yarn, thread, and fabric mills
\item \verb{[150]} miscellaneous textile mill products
\item \verb{[151]} apparel and accessories, except knit
\item \verb{[152]} miscellaneous fabricated textile products
\item \verb{[160]} pulp, paper, and paperboard mills
\item \verb{[161]} miscellaneous paper and pulp products
\item \verb{[162]} paperboard containers and boxes
\item \verb{[171]} newspaper publishing and printing
\item \verb{[172]} printing, publishing, and allied industries, except newspapers
\item \verb{[180]} plastics, synthetics, and resins
\item \verb{[181]} drugs
\item \verb{[182]} soaps and cosmetics
\item \verb{[185]} paints, varnishes, and related products
\item \verb{[190]} agricultural chemicals
\item \verb{[191]} industrial and miscellaneous chemicals
\item \verb{[192]} petroleum and coal products
\item \verb{[200]} petroleum refining
\item \verb{[201]} miscellaneous petroleum and coal products
\item \verb{[210]} tires and inner tubes
\item \verb{[211]} other rubber products, and plastics footwear and belting
\item \verb{[212]} miscellaneous plastics products
\item \verb{[220]} leather tanning and finishing
\item \verb{[221]} footwear, except rubber and plastic
\item \verb{[222]} leather products, except footwear
\item \verb{[230]} logging
\item \verb{[231]} sawmills, planning mills, and millwork
\item \verb{[232]} wood buildings and mobile homes
\item \verb{[241]} miscellaneous wood products
\item \verb{[242]} furniture and fixtures
\item \verb{[250]} glass and glass products
\item \verb{[251]} cement, concrete, gypsum, and plaster products
\item \verb{[252]} structural clay products
\item \verb{[261]} pottery and related products
\item \verb{[262]} miscellaneous nonmetallic mineral and stone products
\item \verb{[270]} blast furnaces, steelworks, rolling and finishing mills
\item \verb{[271]} iron and steel foundries
\item \verb{[272]} primary aluminum industries
\item \verb{[280]} other primary metal industries
\item \verb{[281]} cutlery, hand tools, and other hardware
\item \verb{[282]} fabricated structural metal products
\item \verb{[290]} screw machine products
\item \verb{[291]} metal forgings and stampings
\item \verb{[292]} ordnance
\item \verb{[300]} miscellaneous fabricated metal products
\item \verb{[301]} not specified metal industries
\item \verb{[310]} engines and turbines
\item \verb{[311]} farm machinery and equipment
\item \verb{[312]} construction and material handling machines
\item \verb{[320]} metalworking machinery
\item \verb{[321]} office and accounting machines
\item \verb{[322]} electronic computing equipment
\item \verb{[331]} machinery, except electrical, n.e.c.
\item \verb{[332]} not specified machinery
\item \verb{[340]} household appliances
\item \verb{[341]} radio, tv, and communication equipment
\item \verb{[342]} electrical machinery, equipment, and supplies, n.e.c.
\item \verb{[350]} not specified electrical machinery, equipment, and supplies
\item \verb{[351]} motor vehicles and motor vehicle equipment
\item \verb{[352]} aircraft and parts
\item \verb{[360]} ship and boat building and repairing
\item \verb{[361]} railroad locomotives and equipment
\item \verb{[362]} guided missiles, space vehicles, and parts
\item \verb{[370]} cycles and miscellaneous transportation equipment
\item \verb{[371]} scientific and controlling instruments
\item \verb{[372]} optical and health services supplies
\item \verb{[380]} photographic equipment and supplies
\item \verb{[381]} watches, clocks, and clockwork operated devices
\item \verb{[382]} not specified professional equipment
\item \verb{[390]} toys, amusement, and sporting goods
\item \verb{[391]} miscellaneous manufacturing industries
\item \verb{[392]} not specified manufacturing industries
\item \verb{[400]} railroads
\item \verb{[401]} bus service and urban transit
\item \verb{[402]} taxicab service
\item \verb{[410]} trucking service
\item \verb{[411]} warehousing and storage
\item \verb{[412]} u.s. postal service
\item \verb{[420]} water transportation
\item \verb{[421]} air transportation
\item \verb{[422]} pipe lines, except natural gas
\item \verb{[432]} services incidental to transportation
\item \verb{[440]} radio and television broadcasting
\item \verb{[441]} telephone (wire and radio)
\item \verb{[442]} telegraph and miscellaneous communication service
\item \verb{[460]} electric light and power
\item \verb{[461]} gas and steam supply systems
\item \verb{[462]} electric and gas, and other combinations
\item \verb{[470]} water supply and irrigation
\item \verb{[471]} sanitary services
\item \verb{[472]} not specified utilities
\item \verb{[500]} motor vehicles and equipment
\item \verb{[501]} furniture and home furnishings
\item \verb{[502]} lumber and construction materials
\item \verb{[510]} sporting goods, toys, and hobby goods
\item \verb{[511]} metals and minerals, except petroleum
\item \verb{[512]} electrical goods
\item \verb{[521]} hardware, plumbing and heating supplies
\item \verb{[522]} not specified electrical and hardware products
\item \verb{[530]} machinery, equipment, and supplies
\item \verb{[531]} scrap and waste materials
\item \verb{[532]} miscellaneous wholesale, durable goods
\item \verb{[540]} paper and paper products
\item \verb{[541]} drugs, chemicals, and allied products
\item \verb{[542]} apparel, fabrics, and notions
\item \verb{[550]} groceries and related products
\item \verb{[551]} farm products-raw materials
\item \verb{[552]} petroleum products
\item \verb{[560]} alcoholic beverages
\item \verb{[561]} farm supplies
\item \verb{[562]} miscellaneous wholesale, nondurable goods
\item \verb{[571]} not specified wholesale trade
\item \verb{[580]} lumber and building material retailing
\item \verb{[581]} hardware stores
\item \verb{[582]} retail nurseries and garden stores
\item \verb{[590]} mobile home dealers
\item \verb{[591]} department stores
\item \verb{[592]} variety stores
\item \verb{[600]} miscellaneous general merchandise stores
\item \verb{[601]} grocery stores
\item \verb{[602]} dairy products stores
\item \verb{[610]} retail bakeries
\item \verb{[611]} food stores, n.e.c.
\item \verb{[612]} motor vehicle dealers
\item \verb{[620]} auto and home supply stores
\item \verb{[621]} gasoline service stations
\item \verb{[622]} miscellaneous vehicle dealers
\item \verb{[630]} apparel and accessory stores, except shoe
\item \verb{[631]} shoe stores
\item \verb{[632]} furniture and home furnishings stores
\item \verb{[640]} household appliances, tv, and radio stores
\item \verb{[641]} eating and drinking places
\item \verb{[642]} drug stores
\item \verb{[650]} liquor stores
\item \verb{[651]} sporting goods, bicycles, and hobby stores
\item \verb{[652]} book and stationery stores
\item \verb{[660]} jewelry stores
\item \verb{[661]} sewing, needlework, and piece goods stores
\item \verb{[662]} mail order houses
\item \verb{[670]} vending machine operators
\item \verb{[671]} direct selling establishments
\item \verb{[672]} fuel and ice dealers
\item \verb{[681]} retail florists
\item \verb{[682]} miscellaneous retail stores
\item \verb{[691]} not specified retail trade
\item \verb{[700]} banking
\item \verb{[701]} savings and loan associations
\item \verb{[702]} credit agencies, n.e.c.
\item \verb{[710]} security, commodity brokerage, and investment companies
\item \verb{[711]} insurance
\item \verb{[712]} real estate, including real estate-insurance-law offices
\item \verb{[721]} advertising
\item \verb{[722]} services to dwellings and other buildings
\item \verb{[730]} commercial research, development, and testing labs
\item \verb{[731]} personnel supply services
\item \verb{[732]} business management and consulting services
\item \verb{[740]} computer and data processing services
\item \verb{[741]} detective and protective services
\item \verb{[742]} business services, n.e.c.
\item \verb{[750]} automotive services, except repair
\item \verb{[751]} automotive repair shops
\item \verb{[752]} electrical repair shops
\item \verb{[760]} miscellaneous repair services
\item \verb{[761]} private households
\item \verb{[762]} hotels and motels
\item \verb{[770]} lodging places, except hotels and motels
\item \verb{[771]} laundry, cleaning, and garment services
\item \verb{[772]} beauty shops
\item \verb{[780]} barber shops
\item \verb{[781]} funeral service and crematories
\item \verb{[782]} shoe repair shops
\item \verb{[790]} dressmaking shops
\item \verb{[791]} miscellaneous personal services
\item \verb{[800]} theaters and motion pictures
\item \verb{[801]} bowling alleys, billiard and pool parlors
\item \verb{[802]} miscellaneous entertainment and recreation services
\item \verb{[812]} offices of physicians
\item \verb{[820]} offices of dentists
\item \verb{[821]} offices of chiropractors
\item \verb{[822]} offices of optometrists
\item \verb{[830]} offices of health practitioners, n.e.c.
\item \verb{[831]} hospitals
\item \verb{[832]} nursing and personal care facilities
\item \verb{[840]} health services, n.e.c.
\item \verb{[841]} legal services
\item \verb{[842]} elementary and secondary schools
\item \verb{[850]} colleges and universities
\item \verb{[851]} business, trade, and vocational schools
\item \verb{[852]} libraries
\item \verb{[860]} educational services, n.e.c.
\item \verb{[861]} job training and vocational rehabilitation services
\item \verb{[862]} child day care services
\item \verb{[870]} residential care facilities, without nursing
\item \verb{[871]} social services, n.e.c.
\item \verb{[872]} museums, art galleries and zoos
\item \verb{[880]} religious organizations
\item \verb{[881]} membership organizations
\item \verb{[882]} engineering, architectural, and surveying services
\item \verb{[890]} accounting, auditing, and bookkeeping services
\item \verb{[891]} noncommercial educational and scientific research
\item \verb{[892]} miscellaneous professional and related services
\item \verb{[900]} executive and legislative offices
\item \verb{[901]} general government, n.e.c.
\item \verb{[910]} justice, public order, and safety
\item \verb{[921]} public finance, taxation, and monetary policy
\item \verb{[922]} administration of human resources programs
\item \verb{[930]} administration of environmental quality and household programs
\item \verb{[931]} administration of economic programs
\item \verb{[932]} national security and international affairs
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abhave3}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{ablegal}},
\code{\link{abnomore}},
\code{\link{aborct}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{abspno}},
\code{\link{adoptus}},
\code{\link{adults}},
\code{\link{affrmact}},
\code{\link{age}},
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{agewed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{antiwar}},
\code{\link{arrest}},
\code{\link{attend}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{barate}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blksimp}},
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{bored}},
\code{\link{born}},
\code{\link{brass}},
\code{\link{broadband}},
\code{\link{burglr}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{capfirm}},
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{cathtemp}},
\code{\link{childs}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chlddth}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldsoon}},
\code{\link{cigweek}},
\code{\link{civic}},
\code{\link{civright}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{cobarate}},
\code{\link{codeg}},
\code{\link{coden}},
\code{\link{coden_1222}},
\code{\link{codipged}},
\code{\link{coeduc}},
\code{\link{coeftotlt}},
\code{\link{coevwork}},
\code{\link{cofund}},
\code{\link{cogradtounder}},
\code{\link{cohrs1}},
\code{\link{cohrs2}},
\code{\link{coind10}},
\code{\link{cojail}},
\code{\link{cojew}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{color}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{commute}},
\code{\link{comprend}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlabory}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conorelgsp}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{contemp}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{convictd}},
\code{\link{coocc10}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{coother}},
\code{\link{copres10}},
\code{\link{copunish}},
\code{\link{corel}},
\code{\link{cosector}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{defwrkev}},
\code{\link{defwrknw}},
\code{\link{degree}},
\code{\link{dejavu}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{dotdata}},
\code{\link{dotged}},
\code{\link{dotpeop}},
\code{\link{dotpres}},
\code{\link{dotsvp}},
\code{\link{dotthng}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{earnrs}},
\code{\link{educ}},
\code{\link{educop}},
\code{\link{eftotlt}},
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{era}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{esp}},
\code{\link{evsmoke}},
\code{\link{evwork}},
\code{\link{fair}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{family16sex}},
\code{\link{father}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{febear}},
\code{\link{febrass}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fedirty}},
\code{\link{fefam}},
\code{\link{fefight}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{fehelp}},
\code{\link{fehlpmil}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{femilop}},
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{fenumok}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrite}},
\code{\link{fightair}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{form}},
\code{\link{found}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govcare}},
\code{\link{grace}},
\code{\link{gradtounder}},
\code{\link{granborn}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{guerilla}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}},
\code{\link{gunner}},
\code{\link{gunnum}},
\code{\link{hapcohab}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{healer}},
\code{\link{health}},
\code{\link{helpblk}},
\code{\link{helpful}},
\code{\link{helpnot}},
\code{\link{helpoth}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitnum}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{id}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{immcrime}},
\code{\link{immjobs}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imppol}},
\code{\link{imprelax}},
\code{\link{impwork}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus07}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewtemp}},
\code{\link{jobaccmp}},
\code{\link{jobfind}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{jobtrain}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kidssol}},
\code{\link{king}},
\code{\link{landwar}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{lessnuke}},
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{liberatr}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{libtemp}},
\code{\link{life}},
\code{\link{likediff}},
\code{\link{lockedup}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madeath}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{maind10}},
\code{\link{maindstatus}},
\code{\link{maindtag}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{maocctag}},
\code{\link{mapa}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{mapres80}},
\code{\link{marital}},
\code{\link{marsame}},
\code{\link{marsame1}},
\code{\link{marsamey}},
\code{\link{martype}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{mawrkslf}},
\code{\link{mechanic}},
\code{\link{meserve}},
\code{\link{migage}},
\code{\link{mighist}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{mindbody}},
\code{\link{mininc}},
\code{\link{minmilop}},
\code{\link{mobile16}},
\code{\link{mobiledata}},
\code{\link{morenuke}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{neargod}},
\code{\link{news}()},
\code{\link{nonuke}},
\code{\link{norelgsp16}},
\code{\link{nukewar}},
\code{\link{nurse}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{oth16}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{owngun}},
\code{\link{paattend}},
\code{\link{padeath}},
\code{\link{padeg}},
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{paeduc}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{palefull}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres10}},
\code{\link{papres105plus}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{parborn}},
\code{\link{parsol}},
\code{\link{partfull}},
\code{\link{partyid}},
\code{\link{pawrkgrw}},
\code{\link{pawrkslf}},
\code{\link{perla}},
\code{\link{phone}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{pistol}},
\code{\link{polabuse}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{polhitok}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{polmurdr}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{popespks}},
\code{\link{popular}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{premarsx}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prestg10}},
\code{\link{prestg105plus}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{privacy}},
\code{\link{prottemp}},
\code{\link{prowar}},
\code{\link{quitsmk}},
\code{\link{racavoid}},
\code{\link{raccare}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdif5}},
\code{\link{racdif5x}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs15}},
\code{\link{raceacs2}},
\code{\link{raceacs4}},
\code{\link{raceacs6}},
\code{\link{raceacs8}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{racjoin}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{racthink}},
\code{\link{ractrust}},
\code{\link{racwrite}},
\code{\link{radiohrs}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{reg16}},
\code{\link{reg16_7222}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{relgrade}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{res16}},
\code{\link{resdefwk}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{restact}},
\code{\link{richwork}},
\code{\link{rifle}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{robbry}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{rowngun}},
\code{\link{rushed}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfin}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{satjob}},
\code{\link{school}},
\code{\link{sector}},
\code{\link{sex}},
\code{\link{sexeduc}},
\code{\link{sexnow2}},
\code{\link{shotgun}},
\code{\link{sibdeath}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{size}},
\code{\link{smoke}},
\code{\link{smokecig}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{spanking}},
\code{\link{spattend}},
\code{\link{spbarate}},
\code{\link{spdeath}},
\code{\link{spdeg}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spdipged}},
\code{\link{spdotdat}},
\code{\link{spdotged}},
\code{\link{spdotpeo}},
\code{\link{spdotpre}},
\code{\link{spdotsvp}},
\code{\link{spdotthn}},
\code{\link{spdue}},
\code{\link{speduc}},
\code{\link{speftotlt}},
\code{\link{spevwork}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spgradtounder}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spirits}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{spnorelgsp}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sppaid}},
\code{\link{sppres}},
\code{\link{sppres10}},
\code{\link{sppres105plus}},
\code{\link{sppres80}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{spsector}},
\code{\link{spwrkslf}},
\code{\link{spwrksta}},
\code{\link{srcbelt}},
\code{\link{stresswk}},
\code{\link{strike}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{teenpill}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{teensex}},
\code{\link{thnkself}},
\code{\link{ticket}},
\code{\link{tithing}},
\code{\link{transair}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{tvhours}},
\code{\link{typist}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{upgrade}()},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{vetaid}},
\code{\link{vetfam}},
\code{\link{vetfamnw}},
\code{\link{visions}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{weekswrk}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{widowed}},
\code{\link{wirtap}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}},
\code{\link{workhard}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{wrkwayup}},
\code{\link{xmarsex}},
\code{\link{xmovie}},
\code{\link{xmovie1}},
\code{\link{xnorcsiz}},
\code{\link{year}()},
\code{\link{youngen}}

Other Industry: 
\code{\link{fambiz}},
\code{\link{indfirst}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{kdind80}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{sbind80}},
\code{\link{sphedind}},
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{wrkgovt}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{Core}
\concept{Industry}
\concept{Occupation}
\keyword{variable}
