% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{maindstatus}
\alias{maindstatus}
\title{Maind10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
maindstatus
}
\details{
Question MAIND10 Coding Status
}
\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5999/vshow}{GSS Data Explorer page for \code{maindstatus}}.

Counts by year:\tabular{llllllllll}{
   year \tab found, iap \tab not found, iap \tab found, no answer \tab found, coded \tab not found, no answer \tab not found, coded \tab not found, no code assigned \tab iap \tab Total \cr
   1972 \tab 702 \tab 911 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1234 \tab 270 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1217 \tab 267 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 976 \tab 514 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1230 \tab 269 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 817 \tab 713 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1518 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1293 \tab 175 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1696 \tab 164 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1475 \tab 124 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1458 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1296 \tab 238 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1425 \tab 45 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1511 \tab 308 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1460 \tab 21 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1420 \tab 117 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1499 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1545 \tab 61 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1266 \tab 30 \tab 58 \tab 1614 \tab 3 \tab 17 \tab 4 \tab - \tab 2992 \cr
   1996 \tab 1115 \tab 9 \tab 48 \tab 1720 \tab 1 \tab 8 \tab 3 \tab - \tab 2904 \cr
   1998 \tab 1162 \tab 20 \tab 37 \tab 1602 \tab 1 \tab 9 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 1148 \tab - \tab 44 \tab 1625 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1104 \tab - \tab 42 \tab 1619 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1050 \tab - \tab 79 \tab 1683 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2636 \tab - \tab 60 \tab 1814 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 708 \tab - \tab 32 \tab 1283 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 640 \tab - \tab 41 \tab 1363 \tab - \tab - \tab - \tab - \tab 2044 \cr
   Total \tab 35959 \tab 4317 \tab 441 \tab 14323 \tab 5 \tab 34 \tab 8 \tab 0 \tab 55087 \cr
}
}

\seealso{
Other Core: 
\code{\link{childs}},
\code{\link{maind10}},
\code{\link{maindtag}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{maocctag}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{sibs}}
}
\concept{Core}
\keyword{variable}
