% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{maisco88}
\alias{maisco88}
\title{R's mothers occupation, 1980 census & 1988 isco code}
\source{
General Social Survey https://gss.norc.org
}
\description{
maisco88
}
\details{
Question Respondent's Mother's Occupation, 1988 Census
}
\section{Values}{

\itemize{
\item \verb{[9997]} uncodeable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{isco08}},
\code{\link{isco88}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{paisco08}},
\code{\link{paisco88}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco88}}
}
\concept{Occupation}
\keyword{variable}
