% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{majew}
\alias{majew}
\title{What sort of jew was r's mother}
\source{
General Social Survey https://gss.norc.org
}
\description{
majew
}
\details{
Question B. Was she Orthodox, Conservative, Reform, or none of these?
}
\section{Values}{

\itemize{
\item \verb{[1]} orthodox
\item \verb{[2]} conservative
\item \verb{[3]} reform
\item \verb{[4]} none of these
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1061/vshow}{GSS Data Explorer page for \code{majew}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab orthodox \tab conservative \tab reform \tab none of these \tab don't know \tab no answer \tab Total \cr
   1988 \tab 1444 \tab 8 \tab 14 \tab 7 \tab 1 \tab 1 \tab 6 \tab 1481 \cr
   2008 \tab 1983 \tab 7 \tab 14 \tab 12 \tab 6 \tab 1 \tab - \tab 2023 \cr
   2012 \tab 1945 \tab 5 \tab 5 \tab 14 \tab 5 \tab - \tab - \tab 1974 \cr
   Total \tab 5372 \tab 20 \tab 33 \tab 33 \tab 12 \tab 2 \tab 6 \tab 5478 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
