% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{major2}
\alias{major2}
\title{College major 2}
\source{
General Social Survey https://gss.norc.org
}
\description{
major2
}
\details{
Question What was your major or field of study when you received your (RESPONDENT'S COLLEGE DEGREE) degree? IF THE RESPONDENT RECEIVED MORE THAN ONE GRADUATE LEVEL DEGREE, ASK ABOUT THE HIGHEST DEGREE OBTAINED (E.G. ABOUT A Ph.D. RATHER THAN AN M.B.A).
(COUNT UP TO 2 MENTIONS)
}
\section{Values}{

\itemize{
\item \verb{[1]} agriculture
\item \verb{[2]} environment and natural resources
\item \verb{[3]} architecture
\item \verb{[4]} area, ethnic, and civilization studies
\item \verb{[5]} communications
\item \verb{[6]} communication technologies
\item \verb{[7]} computer and information sciences
\item \verb{[8]} cosmetology services and culinary arts
\item \verb{[9]} education administration and teaching
\item \verb{[10]} engineering
\item \verb{[11]} engineering technologies
\item \verb{[12]} linguistics and foreign languages
\item \verb{[13]} family and consumer sciences
\item \verb{[14]} law and legal studies
\item \verb{[15]} english language, literature, and composition
\item \verb{[16]} liberal arts and humanities
\item \verb{[17]} library science
\item \verb{[18]} biology and life sciences
\item \verb{[19]} mathematics and statistics
\item \verb{[20]} military technologies
\item \verb{[21]} interdisciplinary and multi-disciplinary studies (general)
\item \verb{[22]} physical fitness, parks, recreation, and leisure
\item \verb{[23]} philosophy and religious studies
\item \verb{[24]} theology and religious vocations
\item \verb{[25]} physical sciences
\item \verb{[26]} nuclear, industrial radiology, and biological technologies
\item \verb{[27]} psychology
\item \verb{[28]} criminal justice and fire protection
\item \verb{[29]} public affairs, policy, and social work
\item \verb{[30]} social sciences
\item \verb{[31]} construction services
\item \verb{[32]} electrical and mechanic repairs and technologies
\item \verb{[33]} precision production and industrial arts
\item \verb{[34]} transportation sciences and technologies
\item \verb{[35]} fine arts
\item \verb{[36]} medical and health sciences and services
\item \verb{[37]} business
\item \verb{[38]} history
\item \verb{[39]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/64/vshow}{GSS Data Explorer page for \code{major2}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab environment and natural resources \tab computer and information sciences \tab education administration and teaching \tab engineering \tab engineering technologies \tab linguistics and foreign languages \tab law and legal studies \tab english language, literature, and composition \tab biology and life sciences \tab mathematics and statistics \tab physical fitness, parks, recreation, and leisure \tab philosophy and religious studies \tab theology and religious vocations \tab physical sciences \tab psychology \tab public affairs, policy, and social work \tab social sciences \tab fine arts \tab medical and health sciences and services \tab business \tab history \tab other \tab don't know \tab iap \tab no answer \tab communications \tab library science \tab interdisciplinary and multi-disciplinary studies (general) \tab criminal justice and fire protection \tab agriculture \tab area, ethnic, and civilization studies \tab cosmetology services and culinary arts \tab liberal arts and humanities \tab electrical and mechanic repairs and technologies \tab architecture \tab family and consumer sciences \tab skipped on web \tab military technologies \tab nuclear, industrial radiology, and biological technologies \tab construction services \tab transportation sciences and technologies \tab Total \cr
   2012 \tab - \tab 1 \tab 5 \tab 10 \tab 1 \tab 1 \tab 3 \tab 2 \tab 6 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 4 \tab 6 \tab 2 \tab 19 \tab 8 \tab 9 \tab 25 \tab 3 \tab 5 \tab 2 \tab 1852 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 4 \tab 5 \tab 22 \tab 4 \tab - \tab 5 \tab 3 \tab 4 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 4 \tab 10 \tab 5 \tab 17 \tab 16 \tab 18 \tab 42 \tab 6 \tab 6 \tab 1 \tab 2345 \tab 1 \tab 6 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2 \tab 9 \tab 25 \tab 5 \tab - \tab 1 \tab 1 \tab 5 \tab 7 \tab 9 \tab - \tab - \tab 3 \tab 6 \tab 11 \tab 8 \tab 26 \tab 12 \tab 22 \tab 42 \tab 4 \tab 6 \tab - \tab 2643 \tab - \tab 6 \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 1 \tab 11 \tab 16 \tab - \tab - \tab 3 \tab 2 \tab 7 \tab 4 \tab 11 \tab - \tab - \tab 2 \tab 6 \tab - \tab 3 \tab 10 \tab 5 \tab 8 \tab 28 \tab 8 \tab 8 \tab - \tab 2201 \tab - \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab - \tab 3 \tab 12 \tab 53 \tab 9 \tab 6 \tab 15 \tab 6 \tab 25 \tab 14 \tab 12 \tab - \tab 2 \tab 8 \tab 20 \tab 28 \tab 22 \tab 51 \tab 34 \tab 37 \tab 102 \tab 14 \tab 12 \tab 15 \tab 2941 \tab 30 \tab 19 \tab 1 \tab 1 \tab 13 \tab 3 \tab 5 \tab 3 \tab 6 \tab - \tab 1 \tab 1 \tab 20 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 0 \tab 11 \tab 42 \tab 126 \tab 19 \tab 7 \tab 27 \tab 14 \tab 47 \tab 29 \tab 39 \tab 1 \tab 4 \tab 16 \tab 40 \tab 55 \tab 40 \tab 123 \tab 75 \tab 94 \tab 239 \tab 35 \tab 37 \tab 18 \tab 11982 \tab 32 \tab 35 \tab 6 \tab 2 \tab 24 \tab 5 \tab 7 \tab 4 \tab 12 \tab 2 \tab 1 \tab 1 \tab 20 \tab 0 \tab 0 \tab 0 \tab 0 \tab 13271 \cr
}
}

\seealso{
Other Core: 
\code{\link{codipged}},
\code{\link{cosector}},
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}

Other Education: 
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\concept{Education}
\keyword{variable}
