% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{malekids}
\alias{malekids}
\title{Number of male children}
\source{
General Social Survey https://gss.norc.org
}
\description{
malekids
}
\details{
Question Among your biological children, how many children are male?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7957/vshow}{GSS Data Explorer page for \code{malekids}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 606 \tab 352 \tab 173 \tab 46 \tab 20 \tab 3 \tab 2 \tab 33 \tab 3544 \cr
   Total \tab 2309 \tab 606 \tab 352 \tab 173 \tab 46 \tab 20 \tab 3 \tab 2 \tab 33 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{fluvax}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
