% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{maleornt}
\alias{maleornt}
\title{Interviewer coded male sexual orientation}
\source{
General Social Survey https://gss.norc.org
}
\description{
maleornt
}
\details{
Question INTERVIEWER ONLY, DO NOT ASK R
}
\section{Values}{

\itemize{
\item \code{1} r is homosexual/gay
\item \code{2} r is not homosexual/gay
\item \code{3} r sexual orientation uncertain
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7294/vshow}{GSS Data Explorer page for \code{maleornt}}.
}

\keyword{variable}
