% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{malive}
\alias{malive}
\title{Mother still alive?}
\source{
General Social Survey https://gss.norc.org
}
\description{
malive
}
\details{
Question Is your mother still alive?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
   Full \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4127/vshow}{GSS Data Explorer page for \code{malive}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1986 \tab - \tab 895 \tab 570 \tab 2 \tab 3 \tab 1470 \cr
   1994 \tab 1545 \tab 872 \tab 524 \tab 1 \tab 50 \tab 2992 \cr
   Total \tab 1545 \tab 1767 \tab 1094 \tab 3 \tab 53 \tab 4462 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{macall}},
\code{\link{matime}},
\code{\link{mavisit}},
\code{\link{palive}},
\code{\link{peocntct}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
