% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_53.R
\name{manners}
\alias{manners}
\title{Good manners}
\source{
General Social Survey https://gss.norc.org
}
\description{
manners
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
\enumerate{
\item That he has good manners
}
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 most desirable
\item \verb{[2]} 3 most desirable
\item \verb{[3]} not mentioned
\item \verb{[4]} 3 least desirable
\item \verb{[5]} 1 least desirable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1975, 1976, 1978, 1980, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/474/vshow}{GSS Data Explorer page for \code{manners}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 41 \tab 321 \tab 766 \tab 277 \tab 95 \tab 4 \tab - \tab 1504 \cr
   1975 \tab - \tab 52 \tab 330 \tab 786 \tab 254 \tab 59 \tab 9 \tab - \tab 1490 \cr
   1976 \tab - \tab 39 \tab 277 \tab 803 \tab 290 \tab 79 \tab 11 \tab - \tab 1499 \cr
   1978 \tab - \tab 52 \tab 340 \tab 842 \tab 239 \tab 47 \tab 10 \tab 2 \tab 1532 \cr
   1980 \tab 975 \tab 22 \tab 95 \tab 270 \tab 87 \tab 15 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 50 \tab 353 \tab 831 \tab 284 \tab 61 \tab 20 \tab - \tab 1599 \cr
   Total \tab 975 \tab 256 \tab 1716 \tab 4298 \tab 1431 \tab 356 \tab 57 \tab 3 \tab 9092 \cr
}
}

\seealso{
Other Core: 
\code{\link{conarmyy}},
\code{\link{confedy}},
\code{\link{conjudgy}},
\code{\link{conlaboy}},
\code{\link{conlegiy}},
\code{\link{conmediy}},
\code{\link{conpresy}},
\code{\link{consciy}},
\code{\link{contvy}}
}
\concept{Core}
\keyword{variable}
