% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_56.R
\name{mannersz}
\alias{mannersz}
\title{Good manners-version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
mannersz
}
\details{
Question Please look at the qualities listed on this card. (SHOW CARD:
LET R. INSPECT) All of the qualities may be desirable for a child to have, but could you tell me whether the quality is extremely important, very important, fairly important, not too important, or not at all important. READ AND RECORD RESPONSES FOR EACH.
A. That he has good manners
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely impt
\item \verb{[2]} very impt
\item \verb{[3]} fairly impt
\item \verb{[4]} not too impt
\item \verb{[5]} not at all impt
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/500/vshow}{GSS Data Explorer page for \code{mannersz}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab extremely impt \tab very impt \tab fairly impt \tab not too impt \tab no answer \tab Total \cr
   1980 \tab 985 \tab 135 \tab 261 \tab 81 \tab 3 \tab 3 \tab 1468 \cr
   Total \tab 985 \tab 135 \tab 261 \tab 81 \tab 3 \tab 3 \tab 1468 \cr
}
}

\seealso{
Other Core: 
\code{\link{cleanz}},
\code{\link{considey}},
\code{\link{honestz}},
\code{\link{interesy}},
\code{\link{judgmenz}},
\code{\link{obeysy}},
\code{\link{responsy}},
\code{\link{studiouy}},
\code{\link{successz}}
}
\concept{Core}
\keyword{variable}
