% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_182.R
\name{manprof}
\alias{manprof}
\title{Professional women's lives improved}
\source{
General Social Survey https://gss.norc.org
}
\description{
manprof
}
\details{
Question We'd like your views on how the women's movement has affected certain groups. For each group I name, please tell me whether you think the women's movement has improved their lives, made their lives worse, or had no effect On their lives.
C. Women with management or professional jobs
}
\section{Values}{

\itemize{
\item \code{1} improved their life
\item \code{2} made their life worse
\item \code{3} had no effect on their life
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1702/vshow}{GSS Data Explorer page for \code{manprof}}.
}

\seealso{
Other Gender: 
\code{\link{children}},
\code{\link{men}},
\code{\link{yourself}}

Other Women: 
\code{\link{children}},
\code{\link{men}},
\code{\link{yourself}}
}
\concept{Gender}
\concept{Women}
\keyword{variable}
