% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{mantap}
\alias{mantap}
\title{Should police tap phone of a suspect?}
\source{
General Social Survey https://gss.norc.org
}
\description{
mantap
}
\details{
Question Now, suppose the tip is about a man without a criminal record. Please circle a number for each question to show if you think police should be allowed without a court order to...
B. Tap his telephone.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3999/vshow}{GSS Data Explorer page for \code{mantap}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{databank}},
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{progtax}},
\code{\link{verdict}}

Other Civil Liberties: 
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}}

Other Privacy: 
\code{\link{databank}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}}
}
\concept{Civil Liberties}
\concept{ISSP Role of Gov't}
\concept{Privacy}
\concept{Wiretapping}
\keyword{variable}
