% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{manvsemp}
\alias{manvsemp}
\title{Relations bw management and employees}
\source{
General Social Survey https://gss.norc.org
}
\description{
manvsemp
}
\details{
Question In general, how would you describe relations in your work place between management and employees?
}
\section{Values}{

\itemize{
\item \verb{[1]} very good
\item \verb{[2]} quite good
\item \verb{[3]} neither good nor bad
\item \verb{[4]} quite bad
\item \verb{[5]} very bad
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2811/vshow}{GSS Data Explorer page for \code{manvsemp}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very good \tab quite good \tab neither good nor bad \tab quite bad \tab very bad \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 646 \tab 595 \tab 384 \tab 95 \tab 38 \tab 20 \tab 18 \tab - \tab 2765 \cr
   2004 \tab 945 \tab 637 \tab 679 \tab 360 \tab 102 \tab 22 \tab 26 \tab 41 \tab - \tab 2812 \cr
   2006 \tab 2776 \tab 565 \tab 615 \tab 372 \tab 77 \tab 40 \tab 42 \tab 23 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 393 \tab 443 \tab 216 \tab 77 \tab 14 \tab 10 \tab 34 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 398 \tab 476 \tab 268 \tab 64 \tab 20 \tab 14 \tab 10 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 474 \tab 514 \tab 295 \tab 99 \tab 16 \tab 13 \tab 37 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 549 \tab 752 \tab 479 \tab 110 \tab 41 \tab 21 \tab 16 \tab 21 \tab 3544 \cr
   Total \tab 9290 \tab 3662 \tab 4074 \tab 2374 \tab 624 \tab 191 \tab 146 \tab 179 \tab 21 \tab 20561 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
