% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{maocc10}
\alias{maocc10}
\title{Mothers census occupation code (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
maocc10
}
\details{
Question A. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up?  That is, what was her job called?
OCCUPATION: (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What did she actually do in that job? Tell me, what were some of her main duties?
C. What kind of place did she work for?
INDUSTRY: (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was she self-employed, or did she work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[10]} chief executives
\item \verb{[20]} general and operations managers
\item \verb{[30]} legislators
\item \verb{[40]} advertising and promotions managers
\item \verb{[50]} marketing and sales managers
\item \verb{[60]} public relations and fundraising managers
\item \verb{[100]} administrative services managers
\item \verb{[110]} computer and information systems managers
\item \verb{[120]} financial managers
\item \verb{[135]} compensation and benefits managers
\item \verb{[136]} human resources managers
\item \verb{[137]} training and development managers
\item \verb{[140]} industrial production managers
\item \verb{[150]} purchasing managers
\item \verb{[160]} transportation, storage, and distribution managers
\item \verb{[205]} farmers, ranchers, and other agricultural managers
\item \verb{[220]} construction managers
\item \verb{[230]} education administrators
\item \verb{[300]} architectural and engineering managers
\item \verb{[310]} food service managers
\item \verb{[325]} funeral service managers
\item \verb{[330]} gaming managers
\item \verb{[340]} lodging managers
\item \verb{[350]} medical and health services managers
\item \verb{[360]} natural sciences managers
\item \verb{[400]} postmasters and mail superintendents
\item \verb{[410]} property, real estate, and community association managers
\item \verb{[420]} social and community service managers
\item \verb{[425]} emergency management directors
\item \verb{[430]} managers, all other
\item \verb{[500]} agents and business managers of artists, performers,
\item \verb{[510]} buyers and purchasing agents, farm products
\item \verb{[520]} wholesale and retail buyers, except farm products
\item \verb{[530]} purchasing agents, except wholesale, retail, and farm products
\item \verb{[540]} claims adjusters, appraisers, examiners, and investigators
\item \verb{[565]} compliance officers
\item \verb{[600]} cost estimators
\item \verb{[630]} human resources workers
\item \verb{[640]} compensation, benefits, and job analysis specialists
\item \verb{[650]} training and development specialists
\item \verb{[700]} logisticians
\item \verb{[710]} management analysts
\item \verb{[725]} meeting, convention, and event planners
\item \verb{[726]} fundraisers
\item \verb{[735]} market research analysts and marketing specialists
\item \verb{[740]} business operations specialists, all other
\item \verb{[800]} accountants and auditors
\item \verb{[810]} appraisers and assessors of real estate
\item \verb{[820]} budget analysts
\item \verb{[830]} credit analysts
\item \verb{[840]} financial analysts
\item \verb{[850]} personal financial advisors
\item \verb{[860]} insurance underwriters
\item \verb{[900]} financial examiners
\item \verb{[910]} credit counselors and loan officers
\item \verb{[930]} tax examiners and collectors, and revenue agents
\item \verb{[940]} tax preparers
\item \verb{[950]} financial specialists, all other
\item \verb{[1005]} computer and information research scientists
\item \verb{[1006]} computer systems analysts
\item \verb{[1007]} information security analysts
\item \verb{[1010]} computer programmers
\item \verb{[1020]} software developers, applications and systems software
\item \verb{[1030]} web developers
\item \verb{[1050]} computer support specialists
\item \verb{[1060]} database administrators
\item \verb{[1105]} network and computer systems administrators
\item \verb{[1106]} computer network architects
\item \verb{[1107]} computer occupations, all other
\item \verb{[1200]} actuaries
\item \verb{[1210]} mathematicians
\item \verb{[1220]} operations research analysts
\item \verb{[1230]} statisticians
\item \verb{[1240]} miscellaneous mathematical science occupations
\item \verb{[1300]} architects, except naval
\item \verb{[1310]} surveyors, cartographers, and photogrammetrists
\item \verb{[1320]} aerospace engineers
\item \verb{[1330]} agricultural engineers
\item \verb{[1340]} biomedical engineers
\item \verb{[1350]} chemical engineers
\item \verb{[1360]} civil engineers
\item \verb{[1400]} computer hardware engineers
\item \verb{[1410]} electrical and electronics engineers
\item \verb{[1420]} environmental engineers
\item \verb{[1430]} industrial engineers, including health and safety
\item \verb{[1440]} marine engineers and naval architects
\item \verb{[1450]} materials engineers
\item \verb{[1460]} mechanical engineers
\item \verb{[1500]} mining and geological engineers, including mining safety engineers
\item \verb{[1510]} nuclear engineers
\item \verb{[1520]} petroleum engineers
\item \verb{[1530]} engineers, all other
\item \verb{[1540]} drafters
\item \verb{[1550]} engineering technicians, except drafters
\item \verb{[1560]} surveying and mapping technicians
\item \verb{[1600]} agricultural and food scientists
\item \verb{[1610]} biological scientists
\item \verb{[1640]} conservation scientists and foresters
\item \verb{[1650]} medical scientists
\item \verb{[1660]} life scientists, all other
\item \verb{[1700]} astronomers and physicists
\item \verb{[1710]} atmospheric and space scientists
\item \verb{[1720]} chemists and materials scientists
\item \verb{[1740]} environmental scientists and geoscientists
\item \verb{[1760]} physical scientists, all other
\item \verb{[1800]} economists
\item \verb{[1815]} survey researchers
\item \verb{[1820]} psychologists
\item \verb{[1830]} sociologists
\item \verb{[1840]} urban and regional planners
\item \verb{[1860]} miscellaneous social scientists and related workers
\item \verb{[1900]} agricultural and food science technicians
\item \verb{[1910]} biological technicians
\item \verb{[1920]} chemical technicians
\item \verb{[1930]} geological and petroleum technicians
\item \verb{[1940]} nuclear technicians
\item \verb{[1950]} social science research assistants
\item \verb{[1965]} miscellaneous life, physical, and social science technicians
\item \verb{[2000]} counselors
\item \verb{[2010]} social workers
\item \verb{[2015]} probation officers and correctional treatment specialists
\item \verb{[2016]} social and human service assistants
\item \verb{[2025]} miscellaneous community and social service specialists,
\item \verb{[2040]} clergy
\item \verb{[2050]} directors, religious activities and education
\item \verb{[2060]} religious workers, all other
\item \verb{[2100]} lawyers
\item \verb{[2105]} judicial law clerks
\item \verb{[2110]} judges, magistrates, and other judicial workers
\item \verb{[2145]} paralegals and legal assistants
\item \verb{[2160]} miscellaneous legal support workers
\item \verb{[2200]} postsecondary teachers
\item \verb{[2300]} preschool and kindergarten teachers
\item \verb{[2310]} elementary and middle school teachers
\item \verb{[2320]} secondary school teachers
\item \verb{[2330]} special education teachers
\item \verb{[2340]} other teachers and instructors
\item \verb{[2400]} archivists, curators, and museum technicians
\item \verb{[2430]} librarians
\item \verb{[2440]} library technicians
\item \verb{[2540]} teacher assistants
\item \verb{[2550]} other education, training, and library workers
\item \verb{[2600]} artists and related workers
\item \verb{[2630]} designers
\item \verb{[2700]} actors
\item \verb{[2710]} producers and directors
\item \verb{[2720]} athletes, coaches, umpires, and related workers
\item \verb{[2740]} dancers and choreographers
\item \verb{[2750]} musicians, singers, and related workers
\item \verb{[2760]} entertainers and performers, sports and related workers, all other
\item \verb{[2800]} announcers
\item \verb{[2810]} news analysts, reporters and correspondents
\item \verb{[2825]} public relations specialists
\item \verb{[2830]} editors
\item \verb{[2840]} technical writers
\item \verb{[2850]} writers and authors
\item \verb{[2860]} miscellaneous media and communication  workers
\item \verb{[2900]} broadcast and sound engineering technicians and radio
\item \verb{[2910]} photographers
\item \verb{[2920]} television, video, and motion picture camera operators and
\item \verb{[2960]} media and communication  equipment workers, all other
\item \verb{[3000]} chiropractors
\item \verb{[3010]} dentists
\item \verb{[3030]} dietitians and nutritionists
\item \verb{[3040]} optometrists
\item \verb{[3050]} pharmacists
\item \verb{[3060]} physicians and surgeons
\item \verb{[3110]} physician assistants
\item \verb{[3120]} podiatrists
\item \verb{[3140]} audiologists
\item \verb{[3150]} occupational therapists
\item \verb{[3160]} physical therapists
\item \verb{[3200]} radiation therapists
\item \verb{[3210]} recreational therapists
\item \verb{[3220]} respiratory therapists
\item \verb{[3230]} speech-language  pathologists
\item \verb{[3235]} exercise physiologists
\item \verb{[3245]} therapists, all other
\item \verb{[3250]} veterinarians
\item \verb{[3255]} registered nurses
\item \verb{[3256]} nurse anesthetists
\item \verb{[3257]} nurse midwives
\item \verb{[3258]} nurse practitioners
\item \verb{[3260]} health diagnosing and treating practitioners, all other
\item \verb{[3300]} clinical laboratory technologists and technicians
\item \verb{[3310]} dental hygienists
\item \verb{[3320]} diagnostic related technologists and technicians
\item \verb{[3400]} emergency medical technicians and paramedics
\item \verb{[3420]} health practitioner support technologists and technicians
\item \verb{[3500]} licensed practical and licensed vocational nurses
\item \verb{[3510]} medical records and health information technicians
\item \verb{[3520]} opticians, dispensing
\item \verb{[3535]} miscellaneous health technologists and technicians
\item \verb{[3540]} other healthcare practitioners and technical occupations
\item \verb{[3600]} nursing, psychiatric, and home health aides
\item \verb{[3610]} occupational therapy assistants and aides
\item \verb{[3620]} physical therapist assistants and aides
\item \verb{[3630]} massage therapists
\item \verb{[3640]} dental assistants
\item \verb{[3645]} medical assistants
\item \verb{[3646]} medical transcriptionists
\item \verb{[3647]} pharmacy aides
\item \verb{[3648]} veterinary assistants and laboratory animal caretakers
\item \verb{[3649]} phlebotomists
\item \verb{[3655]} miscellaneous healthcare support occupations, including medical equipment preparers
\item \verb{[3700]} first-line supervisors of correctional officers
\item \verb{[3710]} first-line supervisors of police and detectives
\item \verb{[3720]} first-line supervisors of fire fighting and prevention workers
\item \verb{[3730]} first-line supervisors of protective service workers, all other
\item \verb{[3740]} firefighters
\item \verb{[3750]} fire inspectors
\item \verb{[3800]} bailiffs, correctional officers, and jailers
\item \verb{[3820]} detectives and criminal investigators
\item \verb{[3830]} fish and game wardens
\item \verb{[3840]} parking enforcement workers
\item \verb{[3850]} police and sheriff's patrol officers
\item \verb{[3860]} transit and railroad police
\item \verb{[3900]} animal control workers
\item \verb{[3910]} private detectives and investigators
\item \verb{[3930]} security guards and gaming surveillance officers
\item \verb{[3940]} crossing guards
\item \verb{[3945]} transportation security screeners
\item \verb{[3955]} lifeguards and other recreational, and all other protective service workers
\item \verb{[4000]} chefs and head cooks
\item \verb{[4010]} first-line supervisors of food preparation and serving workers
\item \verb{[4020]} cooks
\item \verb{[4030]} food preparation workers
\item \verb{[4040]} bartenders
\item \verb{[4050]} combined food preparation and serving workers, including fast food
\item \verb{[4060]} counter attendants, cafeteria, food concession, and coffee shop
\item \verb{[4110]} waiters and waitresses
\item \verb{[4120]} food servers, nonrestaurant
\item \verb{[4130]} dining room and cafeteria attendants and bartender helpers
\item \verb{[4140]} dishwashers
\item \verb{[4150]} hosts and hostesses, restaurant, lounge, and coffee shop
\item \verb{[4160]} food preparation and serving related workers, all other
\item \verb{[4200]} first-line supervisors of housekeeping and janitorial workers
\item \verb{[4210]} first-line supervisors of landscaping, lawn service, and groundskeeping workers
\item \verb{[4220]} janitors and building cleaners
\item \verb{[4230]} maids and housekeeping cleaners
\item \verb{[4240]} pest control workers
\item \verb{[4250]} grounds maintenance workers
\item \verb{[4300]} first-line supervisors of gaming workers
\item \verb{[4320]} first-line supervisors of personal service workers
\item \verb{[4340]} animal trainers
\item \verb{[4350]} nonfarm animal caretakers
\item \verb{[4400]} gaming services workers
\item \verb{[4410]} motion picture projectionists
\item \verb{[4420]} ushers, lobby attendants, and ticket takers
\item \verb{[4430]} miscellaneous entertainment attendants and related workers
\item \verb{[4460]} embalmers and funeral attendants
\item \verb{[4465]} morticians, undertakers, and funeral directors
\item \verb{[4500]} barbers
\item \verb{[4510]} hairdressers, hairstylists, and cosmetologists
\item \verb{[4520]} miscellaneous personal appearance workers
\item \verb{[4530]} baggage porters, bellhops, and concierges
\item \verb{[4540]} tour and travel guides
\item \verb{[4600]} childcare workers
\item \verb{[4610]} personal care aides
\item \verb{[4620]} recreation and fitness workers
\item \verb{[4640]} residential advisors
\item \verb{[4650]} personal care and service workers, all other
\item \verb{[4700]} first-line supervisors of retail sales workers
\item \verb{[4710]} first-line supervisors of non-retail sales workers
\item \verb{[4720]} cashiers
\item \verb{[4740]} counter and rental clerks
\item \verb{[4750]} parts salespersons
\item \verb{[4760]} retail salespersons
\item \verb{[4800]} advertising sales agents
\item \verb{[4810]} insurance sales agents
\item \verb{[4820]} securities, commodities, and financial services sales agents
\item \verb{[4830]} travel agents
\item \verb{[4840]} sales representatives, services, all other
\item \verb{[4850]} sales representatives,  wholesale and manufacturing
\item \verb{[4900]} models, demonstrators, and product promoters
\item \verb{[4920]} real estate brokers and sales agents
\item \verb{[4930]} sales engineers
\item \verb{[4940]} telemarketers
\item \verb{[4950]} door-to-door sales workers, news and street vendors, and related workers
\item \verb{[4965]} sales and related workers, all other
\item \verb{[5000]} first-line supervisors of office and administrative support workers
\item \verb{[5010]} switchboard operators, including answering service
\item \verb{[5020]} telephone operators
\item \verb{[5030]} communications  equipment operators, all other
\item \verb{[5100]} bill and account collectors
\item \verb{[5110]} billing and posting clerks
\item \verb{[5120]} bookkeeping, accounting, and auditing clerks
\item \verb{[5130]} gaming cage workers
\item \verb{[5140]} payroll and timekeeping clerks
\item \verb{[5150]} procurement clerks
\item \verb{[5160]} tellers
\item \verb{[5165]} financial clerks, all other
\item \verb{[5200]} brokerage clerks
\item \verb{[5210]} correspondence clerks
\item \verb{[5220]} court, municipal, and license clerks
\item \verb{[5230]} credit authorizers, checkers, and clerks
\item \verb{[5240]} customer service representatives
\item \verb{[5250]} eligibility interviewers, government programs
\item \verb{[5260]} file clerks
\item \verb{[5300]} hotel, motel, and resort desk clerks
\item \verb{[5310]} interviewers, except eligibility and loan
\item \verb{[5320]} library assistants, clerical
\item \verb{[5330]} loan interviewers and clerks
\item \verb{[5340]} new accounts clerks
\item \verb{[5350]} order clerks
\item \verb{[5360]} human resources assistants, except payroll and timekeeping
\item \verb{[5400]} receptionists and information clerks
\item \verb{[5410]} reservation and transportation ticket agents and travel clerks
\item \verb{[5420]} information and record clerks, all other
\item \verb{[5500]} cargo and freight agents
\item \verb{[5510]} couriers and messengers
\item \verb{[5520]} dispatchers
\item \verb{[5530]} meter readers, utilities
\item \verb{[5540]} postal service clerks
\item \verb{[5550]} postal service mail carriers
\item \verb{[5560]} postal service mail sorters, processors, and processing  machine operators
\item \verb{[5600]} production, planning, and expediting clerks
\item \verb{[5610]} shipping, receiving, and traffic clerks
\item \verb{[5620]} stock clerks and order fillers
\item \verb{[5630]} weighers, measurers, checkers, and samplers, recordkeeping
\item \verb{[5700]} secretaries and administrative assistants
\item \verb{[5800]} computer operators
\item \verb{[5810]} data entry keyers
\item \verb{[5820]} word processors and typists
\item \verb{[5830]} desktop publishers
\item \verb{[5840]} insurance claims and policy processing clerks
\item \verb{[5850]} mail clerks and mail machine operators, except postal service
\item \verb{[5860]} office clerks, general
\item \verb{[5900]} office machine operators, except computer
\item \verb{[5910]} proofreaders and copy markers
\item \verb{[5920]} statistical assistants
\item \verb{[5940]} office and administrative support workers, all other
\item \verb{[6005]} first-line supervisors of farming, fishing, and forestry workers
\item \verb{[6010]} agricultural inspectors
\item \verb{[6020]} animal breeders
\item \verb{[6040]} graders and sorters, agricultural products
\item \verb{[6050]} miscellaneous agricultural workers
\item \verb{[6100]} fishers and related fishing workers
\item \verb{[6110]} hunters and trappers
\item \verb{[6120]} forest and conservation workers
\item \verb{[6130]} logging workers
\item \verb{[6200]} first-line supervisors of construction trades and extraction workers
\item \verb{[6210]} boilermakers
\item \verb{[6220]} brickmasons, blockmasons, and stonemasons
\item \verb{[6230]} carpenters
\item \verb{[6240]} carpet, floor, and tile installers and finishers
\item \verb{[6250]} cement masons, concrete finishers, and terrazzo workers
\item \verb{[6260]} construction laborers
\item \verb{[6300]} paving, surfacing, and tamping equipment operators
\item \verb{[6310]} pile-driver operators
\item \verb{[6320]} operating engineers and other construction equipment  operators
\item \verb{[6330]} drywall installers, ceiling tile installers, and tapers
\item \verb{[6355]} electricians
\item \verb{[6360]} glaziers
\item \verb{[6400]} insulation workers
\item \verb{[6420]} painters, construction and maintenance
\item \verb{[6430]} paperhangers
\item \verb{[6440]} pipelayers, plumbers, pipefitters, and steamfitters
\item \verb{[6460]} plasterers and stucco masons
\item \verb{[6500]} reinforcing iron and rebar workers
\item \verb{[6515]} roofers
\item \verb{[6520]} sheet metal workers
\item \verb{[6530]} structural iron and steel workers
\item \verb{[6540]} solar photovoltaic installers
\item \verb{[6600]} helpers, construction trades
\item \verb{[6660]} construction and building inspectors
\item \verb{[6700]} elevator installers and repairers
\item \verb{[6710]} fence erectors
\item \verb{[6720]} hazardous materials removal workers
\item \verb{[6730]} highway maintenance workers
\item \verb{[6740]} rail-track laying and maintenance equipment operators
\item \verb{[6750]} septic tank servicers and sewer pipe cleaners
\item \verb{[6765]} miscellaneous construction and related workers
\item \verb{[6800]} derrick, rotary drill, and service unit operators, oil, gas, and mining
\item \verb{[6820]} earth drillers, except oil and gas
\item \verb{[6830]} explosives workers, ordnance handling experts, and blasters
\item \verb{[6840]} mining machine operators
\item \verb{[6910]} roof bolters, mining
\item \verb{[6920]} roustabouts, oil and gas
\item \verb{[6930]} helpers-extraction workers
\item \verb{[6940]} other extraction workers
\item \verb{[7000]} first-line supervisors of mechanics, installers, and repairers
\item \verb{[7010]} computer, automated teller, and office machine repairers
\item \verb{[7020]} radio and telecommunications equipment installers and repairers
\item \verb{[7030]} avionics technicians
\item \verb{[7040]} electric motor, power tool, and related repairers
\item \verb{[7050]} electrical and electronics installers and repairers, transportation
\item \verb{[7100]} electrical and electronics repairers, industrial and utility
\item \verb{[7110]} electronic equipment installers and repairers, motor vehicles
\item \verb{[7120]} electronic home entertainment equipment installers and repairers
\item \verb{[7130]} security and fire alarm systems installers
\item \verb{[7140]} aircraft mechanics and service technicians
\item \verb{[7150]} automotive body and related repairers
\item \verb{[7160]} automotive glass installers and repairers
\item \verb{[7200]} automotive service technicians and mechanics
\item \verb{[7210]} bus and truck mechanics and diesel engine specialists
\item \verb{[7220]} heavy vehicle and mobile equipment service technicians and mechanics
\item \verb{[7240]} small engine mechanics
\item \verb{[7260]} miscellaneous vehicle and mobile equipment mechanics, installers, and repairers
\item \verb{[7300]} control and valve installers and repairers
\item \verb{[7315]} heating, air conditioning, and refrigeration mechanics and installers
\item \verb{[7320]} home appliance repairers
\item \verb{[7330]} industrial and refractory machinery mechanics
\item \verb{[7340]} maintenance and repair workers, general
\item \verb{[7350]} maintenance workers, machinery
\item \verb{[7360]} millwrights
\item \verb{[7410]} electrical power-line installers and repairers
\item \verb{[7420]} telecommunications  line installers and repairers
\item \verb{[7430]} precision instrument and equipment repairers
\item \verb{[7440]} wind turbine service technicians
\item \verb{[7510]} coin, vending, and amusement machine servicers and repairers
\item \verb{[7520]} commercial divers
\item \verb{[7540]} locksmiths and safe repairers
\item \verb{[7550]} manufactured building and mobile home installers
\item \verb{[7560]} riggers
\item \verb{[7600]} signal and track switch repairers
\item \verb{[7610]} helpers-installation,  maintenance, and repair workers
\item \verb{[7630]} other installation, maintenance, and repair workers
\item \verb{[7700]} first-line supervisors of production and operating workers
\item \verb{[7710]} aircraft structure, surfaces, rigging, and systems assemblers
\item \verb{[7720]} electrical, electronics, and electromechanical assemblers
\item \verb{[7730]} engine and other machine assemblers
\item \verb{[7740]} structural metal fabricators and fitters
\item \verb{[7750]} miscellaneous assemblers and fabricators
\item \verb{[7800]} bakers
\item \verb{[7810]} butchers and other meat, poultry, and fish processing workers
\item \verb{[7830]} food and tobacco roasting, baking, and drying machine operators and tenders
\item \verb{[7840]} food batchmakers
\item \verb{[7850]} food cooking machine operators and tenders
\item \verb{[7855]} food processing workers, all other
\item \verb{[7900]} computer control programmers and operators
\item \verb{[7920]} extruding and drawing machine setters, operators, and tenders,  metal and plastic
\item \verb{[7930]} forging machine setters, operators, and tenders, metal and plastic
\item \verb{[7940]} rolling machine setters, operators, and tenders, metal and plastic
\item \verb{[7950]} cutting, punching, and press machine setters, operators, and tenders, metal and plastic
\item \verb{[7960]} drilling and boring machine tool setters, operators, and tenders,  metal and plastic
\item \verb{[8000]} grinding, lapping, polishing, and buffing machine tool setters,  operators, and tenders, metal and plastic
\item \verb{[8010]} lathe and turning machine tool setters, operators, and tenders, metal and plastic
\item \verb{[8020]} milling and planing machine setters, operators, and tenders, metal  and plastic
\item \verb{[8030]} machinists
\item \verb{[8040]} metal furnace operators, tenders, pourers, and casters
\item \verb{[8060]} model makers and patternmakers, metal and plastic
\item \verb{[8100]} molders and molding machine setters, operators, and tenders, metal and plastic
\item \verb{[8120]} multiple machine tool setters, operators, and tenders, metal and plastic
\item \verb{[8130]} tool and die makers
\item \verb{[8140]} welding, soldering, and brazing workers
\item \verb{[8150]} heat treating equipment setters, operators, and tenders, metal and plastic
\item \verb{[8160]} layout workers, metal and plastic
\item \verb{[8200]} plating and coating machine setters, operators, and tenders, metal  and plastic
\item \verb{[8210]} tool grinders, filers, and sharpeners
\item \verb{[8220]} metal workers and plastic workers, all other
\item \verb{[8250]} prepress technicians and workers
\item \verb{[8255]} printing press operators
\item \verb{[8256]} print binding and finishing workers
\item \verb{[8300]} laundry and dry-cleaning workers
\item \verb{[8310]} pressers, textile, garment, and related materials
\item \verb{[8320]} sewing machine operators
\item \verb{[8330]} shoe and leather workers and repairers
\item \verb{[8340]} shoe machine operators and tenders
\item \verb{[8350]} tailors, dressmakers, and sewers
\item \verb{[8360]} textile bleaching and dyeing machine operators and tenders
\item \verb{[8400]} textile cutting machine setters, operators, and tenders
\item \verb{[8410]} textile knitting and weaving machine setters, operators, and tenders
\item \verb{[8420]} textile winding, twisting, and drawing out machine setters, operators,  and tenders
\item \verb{[8430]} extruding and forming machine setters, operators, and tenders,  synthetic and glass fibers
\item \verb{[8440]} fabric and apparel patternmakers
\item \verb{[8450]} upholsterers
\item \verb{[8460]} textile, apparel, and furnishings workers, all other
\item \verb{[8500]} cabinetmakers and bench carpenters
\item \verb{[8510]} furniture finishers
\item \verb{[8520]} model makers and patternmakers, wood
\item \verb{[8530]} sawing machine setters, operators, and tenders, wood
\item \verb{[8540]} woodworking machine setters, operators, and tenders, except sawing
\item \verb{[8550]} woodworkers, all other
\item \verb{[8600]} power plant operators, distributors, and dispatchers
\item \verb{[8610]} stationary engineers and boiler operators
\item \verb{[8620]} water and wastewater treatment plant and system operators
\item \verb{[8630]} miscellaneous plant and system operators
\item \verb{[8640]} chemical processing machine setters, operators, and tenders
\item \verb{[8650]} crushing, grinding, polishing, mixing, and blending workers
\item \verb{[8710]} cutting workers
\item \verb{[8720]} extruding, forming, pressing, and compacting machine setters,
\item \verb{[8730]} operators, and tenders furnace, kiln, oven, drier, and kettle operators and tenders
\item \verb{[8740]} inspectors, testers, sorters, samplers, and weighers
\item \verb{[8750]} jewelers and precious stone and metal workers
\item \verb{[8760]} medical, dental, and ophthalmic laboratory technicians
\item \verb{[8800]} packaging and filling machine operators and tenders
\item \verb{[8810]} painting workers
\item \verb{[8830]} photographic process workers and processing machine operators
\item \verb{[8840]} semiconductor processors
\item \verb{[8850]} adhesive bonding machine operators and tenders
\item \verb{[8860]} cleaning, washing, and metal pickling equipment operators and tenders
\item \verb{[8900]} cooling and freezing equipment operators and tenders
\item \verb{[8910]} etchers and engravers
\item \verb{[8920]} molders, shapers, and casters, except metal and plastic
\item \verb{[8930]} paper goods machine setters, operators, and tenders
\item \verb{[8940]} tire builders
\item \verb{[8950]} helpers-production workers
\item \verb{[8965]} production workers, all other
\item \verb{[9000]} supervisors of transportation and material moving workers
\item \verb{[9030]} aircraft pilots and flight engineers
\item \verb{[9040]} air traffic controllers and airfield operations specialists
\item \verb{[9050]} flight attendants
\item \verb{[9110]} ambulance drivers and attendants, except emergency medical  technicians
\item \verb{[9120]} bus drivers
\item \verb{[9130]} driver/sales workers and truck drivers
\item \verb{[9140]} taxi drivers and chauffeurs
\item \verb{[9150]} motor vehicle operators, all other
\item \verb{[9200]} locomotive engineers and operators
\item \verb{[9230]} railroad brake, signal, and switch operators
\item \verb{[9240]} railroad conductors and yardmasters
\item \verb{[9260]} subway, streetcar, and other rail transportation workers
\item \verb{[9300]} sailors and marine oilers
\item \verb{[9310]} ship and boat captains and operators
\item \verb{[9330]} ship engineers
\item \verb{[9340]} bridge and lock tenders
\item \verb{[9350]} parking lot attendants
\item \verb{[9360]} automotive and watercraft service attendants
\item \verb{[9410]} transportation inspectors
\item \verb{[9415]} transportation attendants, except flight attendants
\item \verb{[9420]} other transportation workers
\item \verb{[9500]} conveyor operators and tenders
\item \verb{[9510]} crane and tower operators
\item \verb{[9520]} dredge, excavating, and loading machine operators
\item \verb{[9560]} hoist and winch operators
\item \verb{[9600]} industrial truck and tractor operators
\item \verb{[9610]} cleaners of vehicles and equipment
\item \verb{[9620]} laborers and freight, stock, and material movers, hand
\item \verb{[9630]} machine feeders and offbearers
\item \verb{[9640]} packers and packagers, hand
\item \verb{[9650]} pumping station operators
\item \verb{[9720]} refuse and recyclable material collectors
\item \verb{[9730]} mine shuttle car operators
\item \verb{[9740]} tank car, truck, and ship loaders
\item \verb{[9750]} material moving workers, all other
\item \verb{[9800]} military officer special and tactical operations leaders
\item \verb{[9810]} first-line enlisted military supervisors
\item \verb{[9820]} military enlisted tactical operations and air/weapons specialists and crew members
\item \verb{[9830]} military, rank not specified
\item \verb{[9997]} uncodeable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abhave3}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{ablegal}},
\code{\link{abnomore}},
\code{\link{aborct}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{abspno}},
\code{\link{adoptus}},
\code{\link{adults}},
\code{\link{affrmact}},
\code{\link{age}},
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{agewed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{antiwar}},
\code{\link{arrest}},
\code{\link{attend}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{barate}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blksimp}},
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{bored}},
\code{\link{born}},
\code{\link{brass}},
\code{\link{broadband}},
\code{\link{burglr}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{capfirm}},
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{cathtemp}},
\code{\link{childs}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chlddth}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldsoon}},
\code{\link{cigweek}},
\code{\link{civic}},
\code{\link{civright}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{cobarate}},
\code{\link{codeg}},
\code{\link{coden}},
\code{\link{coden_1222}},
\code{\link{codipged}},
\code{\link{coeduc}},
\code{\link{coeftotlt}},
\code{\link{coevwork}},
\code{\link{cofund}},
\code{\link{cogradtounder}},
\code{\link{cohrs1}},
\code{\link{cohrs2}},
\code{\link{coind10}},
\code{\link{cojail}},
\code{\link{cojew}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{color}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{commute}},
\code{\link{comprend}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlabory}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conorelgsp}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{contemp}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{convictd}},
\code{\link{coocc10}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{coother}},
\code{\link{copres10}},
\code{\link{copunish}},
\code{\link{corel}},
\code{\link{cosector}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{defwrkev}},
\code{\link{defwrknw}},
\code{\link{degree}},
\code{\link{dejavu}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{dotdata}},
\code{\link{dotged}},
\code{\link{dotpeop}},
\code{\link{dotpres}},
\code{\link{dotsvp}},
\code{\link{dotthng}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{earnrs}},
\code{\link{educ}},
\code{\link{educop}},
\code{\link{eftotlt}},
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{era}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{esp}},
\code{\link{evsmoke}},
\code{\link{evwork}},
\code{\link{fair}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{family16sex}},
\code{\link{father}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{febear}},
\code{\link{febrass}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fedirty}},
\code{\link{fefam}},
\code{\link{fefight}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{fehelp}},
\code{\link{fehlpmil}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{femilop}},
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{fenumok}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrite}},
\code{\link{fightair}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{form}},
\code{\link{found}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govcare}},
\code{\link{grace}},
\code{\link{gradtounder}},
\code{\link{granborn}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{guerilla}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}},
\code{\link{gunner}},
\code{\link{gunnum}},
\code{\link{hapcohab}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{healer}},
\code{\link{health}},
\code{\link{helpblk}},
\code{\link{helpful}},
\code{\link{helpnot}},
\code{\link{helpoth}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitnum}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{id}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{immcrime}},
\code{\link{immjobs}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imppol}},
\code{\link{imprelax}},
\code{\link{impwork}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus07}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewtemp}},
\code{\link{jobaccmp}},
\code{\link{jobfind}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{jobtrain}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kidssol}},
\code{\link{king}},
\code{\link{landwar}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{lessnuke}},
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{liberatr}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{libtemp}},
\code{\link{life}},
\code{\link{likediff}},
\code{\link{lockedup}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madeath}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maindstatus}},
\code{\link{maindtag}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{maocc80}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{maocctag}},
\code{\link{mapa}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{mapres80}},
\code{\link{marital}},
\code{\link{marsame}},
\code{\link{marsame1}},
\code{\link{marsamey}},
\code{\link{martype}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{mawrkslf}},
\code{\link{mechanic}},
\code{\link{meserve}},
\code{\link{migage}},
\code{\link{mighist}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{mindbody}},
\code{\link{mininc}},
\code{\link{minmilop}},
\code{\link{mobile16}},
\code{\link{mobiledata}},
\code{\link{morenuke}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{neargod}},
\code{\link{news}()},
\code{\link{nonuke}},
\code{\link{norelgsp16}},
\code{\link{nukewar}},
\code{\link{nurse}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{oth16}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{owngun}},
\code{\link{paattend}},
\code{\link{padeath}},
\code{\link{padeg}},
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{paeduc}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{palefull}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres10}},
\code{\link{papres105plus}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{parborn}},
\code{\link{parsol}},
\code{\link{partfull}},
\code{\link{partyid}},
\code{\link{pawrkgrw}},
\code{\link{pawrkslf}},
\code{\link{perla}},
\code{\link{phone}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{pistol}},
\code{\link{polabuse}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{polhitok}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{polmurdr}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{popespks}},
\code{\link{popular}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{premarsx}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prestg10}},
\code{\link{prestg105plus}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{privacy}},
\code{\link{prottemp}},
\code{\link{prowar}},
\code{\link{quitsmk}},
\code{\link{racavoid}},
\code{\link{raccare}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdif5}},
\code{\link{racdif5x}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs15}},
\code{\link{raceacs2}},
\code{\link{raceacs4}},
\code{\link{raceacs6}},
\code{\link{raceacs8}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{racjoin}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{racthink}},
\code{\link{ractrust}},
\code{\link{racwrite}},
\code{\link{radiohrs}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{reg16}},
\code{\link{reg16_7222}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{relgrade}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{res16}},
\code{\link{resdefwk}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{restact}},
\code{\link{richwork}},
\code{\link{rifle}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{robbry}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{rowngun}},
\code{\link{rushed}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfin}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{satjob}},
\code{\link{school}},
\code{\link{sector}},
\code{\link{sex}},
\code{\link{sexeduc}},
\code{\link{sexnow2}},
\code{\link{shotgun}},
\code{\link{sibdeath}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{size}},
\code{\link{smoke}},
\code{\link{smokecig}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{spanking}},
\code{\link{spattend}},
\code{\link{spbarate}},
\code{\link{spdeath}},
\code{\link{spdeg}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spdipged}},
\code{\link{spdotdat}},
\code{\link{spdotged}},
\code{\link{spdotpeo}},
\code{\link{spdotpre}},
\code{\link{spdotsvp}},
\code{\link{spdotthn}},
\code{\link{spdue}},
\code{\link{speduc}},
\code{\link{speftotlt}},
\code{\link{spevwork}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spgradtounder}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spirits}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{spnorelgsp}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sppaid}},
\code{\link{sppres}},
\code{\link{sppres10}},
\code{\link{sppres105plus}},
\code{\link{sppres80}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{spsector}},
\code{\link{spwrkslf}},
\code{\link{spwrksta}},
\code{\link{srcbelt}},
\code{\link{stresswk}},
\code{\link{strike}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{teenpill}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{teensex}},
\code{\link{thnkself}},
\code{\link{ticket}},
\code{\link{tithing}},
\code{\link{transair}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{tvhours}},
\code{\link{typist}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{upgrade}()},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{vetaid}},
\code{\link{vetfam}},
\code{\link{vetfamnw}},
\code{\link{visions}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{weekswrk}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{widowed}},
\code{\link{wirtap}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}},
\code{\link{workhard}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{wrkwayup}},
\code{\link{xmarsex}},
\code{\link{xmovie}},
\code{\link{xmovie1}},
\code{\link{xnorcsiz}},
\code{\link{year}()},
\code{\link{youngen}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{Core}
\concept{Occupation}
\keyword{variable}
