% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{maocc80}
\alias{maocc80}
\title{Mothers census occupation code (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
maocc80
}
\details{
Question A. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up?  That is, what was her job called?
OCCUPATION: (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What did she actually do in that job? Tell me, what were some of her main duties?
C. What kind of place did she work for?
INDUSTRY: (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was she self-employed, or did she work for someone else?
}
\section{Values}{

\itemize{
\item \code{3} legislators
\item \code{4} chief executives and general admin., public admin
\item \code{5} administrators and officials, public administration
\item \code{6} administrators, protective service
\item \code{7} financial managers
\item \code{8} personnel and labor relations managers
\item \code{9} purchasing managers
\item \code{13} managers, marketing, advertising, and public relations
\item \code{14} administrators, education and related fields
\item \code{15} managers, medicine and health
\item \code{16} managers, properties and real estate
\item \code{17} postmasters and mail superintendents
\item \code{18} funeral directors
\item \code{19} managers and administrators, n.e.c.
\item \code{23} accountants and auditors
\item \code{24} underwriters
\item \code{25} other financial officers
\item \code{26} management analysts
\item \code{27} personnel, training, and labor relations specialists
\item \code{28} purchasing agents and buyers, farm products
\item \code{29} buyers, wholesale and retail trade except farm products
\item \code{33} purchasing agents and buyers
\item \code{34} business and promotion agents
\item \code{35} construction inspectors
\item \code{36} inspectors/compliance officers, except construction
\item \code{37} management related occupations, n.e.c.
\item \code{43} architects
\item \code{44} aerospace engineers
\item \code{45} metallurgical and materials engineers
\item \code{46} mining engineers
\item \code{47} petroleum engineers
\item \code{48} chemical engineers
\item \code{49} nuclear engineers
\item \code{53} civil engineers
\item \code{54} agricultural engineers
\item \code{55} electrical and electronic engineers
\item \code{56} industrial engineers
\item \code{57} mechanical engineers
\item \code{58} marine and naval architects
\item \code{59} engineers, n.e.c.
\item \code{63} surveyors and mapping scientists
\item \code{64} computer systems analysts and scientists
\item \code{65} operations and systems researchers and analysts
\item \code{66} actuaries
\item \code{67} statisticians
\item \code{68} mathematical scientists, n.e.c.
\item \code{69} physicists and astronomers
\item \code{73} chemists, except biochemists
\item \code{74} atmospheric and space scientists
\item \code{75} geologists and geodesists
\item \code{76} physical scientists, n.e.c.
\item \code{77} agricultural and food scientists
\item \code{78} biological and life scientists
\item \code{79} forestry and conservation scientists
\item \code{83} medical scientists
\item \code{84} physicians
\item \code{85} dentists
\item \code{86} veterinarians
\item \code{87} optometrists
\item \code{88} podiatrists
\item \code{89} health diagnosing practitioners, n.e.c.
\item \code{95} registered nurses
\item \code{96} pharmacists
\item \code{97} dietitians
\item \code{98} inhalation therapists
\item \code{99} occupational therapists
\item \code{103} physical therapists
\item \code{104} speech therapists
\item \code{105} therapists, n.e.c.
\item \code{106} physicians' assistants
\item \code{113} earth, environmental, and marine science teachers
\item \code{114} biological science teachers
\item \code{115} chemistry teachers
\item \code{116} physics teachers
\item \code{117} natural science teachers, n.e.c.
\item \code{118} psychology teachers
\item \code{119} economics teachers
\item \code{123} history teachers
\item \code{124} political science teachers
\item \code{125} sociology teachers
\item \code{126} social science teachers, n.e.c.
\item \code{127} engineering teachers
\item \code{128} mathematical science teachers
\item \code{129} computer science teachers
\item \code{133} medical science teachers
\item \code{134} health specialties teachers
\item \code{135} business, commerce, and marketing teachers
\item \code{136} agriculture and forestry teachers
\item \code{137} art, drama, and music teachers
\item \code{138} physical education teachers
\item \code{139} education teachers
\item \code{143} english teachers
\item \code{144} foreign language teachers
\item \code{145} law teachers
\item \code{146} social work teachers
\item \code{147} theology teachers
\item \code{148} trade and industrial teachers
\item \code{149} home economics teachers
\item \code{153} teachers, postsecondary, n.e.c.
\item \code{154} postsecondary teachers, subject not specified
\item \code{155} teachers, prekindergarten and kindergarten
\item \code{156} teachers, elementary school
\item \code{157} teachers, secondary school
\item \code{158} teachers, special education
\item \code{159} teachers, n.e.c.
\item \code{163} counselors, educational and vocational
\item \code{164} librarians
\item \code{165} archivists and curators
\item \code{166} economists
\item \code{167} psychologists
\item \code{168} sociologists
\item \code{169} social scientists, n.e.c.
\item \code{173} urban planners
\item \code{174} social workers
\item \code{175} recreation workers
\item \code{176} clergy
\item \code{177} religious workers, n.e.c.
\item \code{178} lawyers
\item \code{179} judges
\item \code{183} authors
\item \code{184} technical writers
\item \code{185} designers
\item \code{186} musicians and composers
\item \code{187} actors and directors
\item \code{188} painters, sculptors, craft-artists, and printmakers
\item \code{189} photographers
\item \code{193} dancers
\item \code{194} artists, performers, and related workers, n.e.c.
\item \code{195} editors and reporters
\item \code{197} public relations specialists
\item \code{198} announcers
\item \code{199} athletes
\item \code{203} clinical laboratory technologists and technicians
\item \code{204} dental hygienists
\item \code{205} health record technologists and technicians
\item \code{206} radiologic technicians
\item \code{207} licensed practical nurses
\item \code{208} health technologists and technicians, n.e.c.
\item \code{213} electrical and electronic technicians
\item \code{214} industrial engineering technicians
\item \code{215} mechanical engineering technicians
\item \code{216} engineering technicians, n.e.c.
\item \code{217} drafting occupations
\item \code{218} surveying and mapping technicians
\item \code{223} biological technicians
\item \code{224} chemical technicians
\item \code{225} science technicians, n.e.c.
\item \code{226} airplane pilots and navigators
\item \code{227} air traffic controllers
\item \code{228} broadcast equipment operators
\item \code{229} computer programmers
\item \code{233} tool programmers, numerical control
\item \code{234} legal assistants
\item \code{235} technicians, n.e.c.
\item \code{243} supervisors and proprietors, sales occupations
\item \code{253} insurance sales occupations
\item \code{254} real estate sales occupations
\item \code{255} securities and financial services sales occupations
\item \code{256} advertising and related sales occupations
\item \code{257} sales occupations, other business services
\item \code{258} sales engineers
\item \code{259} sales rep., mining, manufacturing, and wholesale
\item \code{263} sales workers, motor vehicles and boats
\item \code{264} sales workers, apparel
\item \code{265} sales workers, shoes
\item \code{266} sales workers, furniture and home furnishings
\item \code{267} sales workers, radio, tv, hi-fi, and appliances
\item \code{268} sales workers, hardware and building supplies
\item \code{269} sales workers, parts
\item \code{274} sales workers, other commodities
\item \code{275} sales counter clerks
\item \code{276} cashiers
\item \code{277} street and door-to-door sales workers
\item \code{278} news vendors
\item \code{283} demonstrators, promoters and models, sales
\item \code{284} auctioneers
\item \code{285} sales support occupations, n.e.c.
\item \code{303} supervisors, general office
\item \code{304} supervisors, computer equipment operators
\item \code{305} supervisors, financial records processing
\item \code{306} chief communications operators
\item \code{307} supervisors, distribution, scheduling, and adjusting clerks
\item \code{308} computer operators
\item \code{309} peripheral equipment operators
\item \code{313} secretaries
\item \code{314} stenographers
\item \code{315} typists
\item \code{316} interviewers
\item \code{317} hotel clerks
\item \code{318} transportation ticket and reservation agents
\item \code{319} receptionists
\item \code{323} information clerks, n.e.c.
\item \code{325} classified-ad clerks
\item \code{326} correspondence clerks
\item \code{327} order clerks
\item \code{328} personnel clerks, except payroll and timekeeping
\item \code{329} library clerks
\item \code{335} file clerks
\item \code{336} records clerks
\item \code{337} bookkeepers, accounting and auditing clerks
\item \code{338} payroll and timekeeping clerks
\item \code{339} billing clerks
\item \code{343} cost and rate clerks
\item \code{344} billing, posting, and calculating machine operators
\item \code{345} duplicating machine operators
\item \code{346} mail preparing and paper handling machine operators
\item \code{347} office machine operators, n.e.c.
\item \code{348} telephone operators
\item \code{349} telegraphers
\item \code{353} communications equipment operators, n.e.c.
\item \code{354} postal clerks, excluding mail carriers
\item \code{355} mail carriers, postal service
\item \code{356} mail clerks, excluding postal service
\item \code{357} messengers
\item \code{359} dispatchers
\item \code{363} production coordinators
\item \code{364} traffic, shipping and receiving clerks
\item \code{365} stock and inventory clerks
\item \code{366} meter readers
\item \code{368} weighers, measurers, and checkers
\item \code{369} samplers
\item \code{373} expediters
\item \code{374} material recording, scheduling and distributing clerks, n.e.c.
\item \code{375} insurance adjusters, examiners, and investigators
\item \code{376} investigators and adjusters, except insurance
\item \code{377} eligibility clerks, social welfare
\item \code{378} bill and account collectors
\item \code{379} general office clerks
\item \code{383} bank tellers
\item \code{384} proofreaders
\item \code{385} data-entry keyers
\item \code{386} statistical clerks
\item \code{387} teachers' aides
\item \code{389} administrative support occupations, n.e.c.
\item \code{403} launderers and ironers
\item \code{404} cooks, private household
\item \code{405} housekeepers and butlers
\item \code{406} child care workers, private household
\item \code{407} private household cleaners and servants
\item \code{413} supervisors, firefighting and fire prevention occupations
\item \code{414} supervisors, police and detectives
\item \code{415} supervisors, guards
\item \code{416} fire inspection and fire prevention occupations
\item \code{417} firefighting occupations
\item \code{418} police and detectives, public service
\item \code{423} sheriffs, bailiffs, and other law enforcement officers
\item \code{424} correctional institution officers
\item \code{425} crossing guards
\item \code{426} guards and police, excluding public service
\item \code{427} protective service occupations
\item \code{430} former member of the armed forces
\item \code{431} current member of the armed forces
\item \code{433} supervisors, food preparation and service occupations
\item \code{434} bartenders
\item \code{435} waiters and waitresses
\item \code{436} cooks, except short order
\item \code{437} short-order cooks
\item \code{438} food counter, fountain and related occupations
\item \code{439} kitchen workers, food preparation
\item \code{443} waiters'/waitresses' assistants
\item \code{444} miscellaneous food preparation occupations
\item \code{445} dental assistants
\item \code{446} health aids, except nursing
\item \code{447} nursing aides, orderlies and attendants
\item \code{448} supervisors, cleaning and building service workers
\item \code{449} maids and housemen
\item \code{453} janitors and cleaners
\item \code{454} elevator operators
\item \code{455} pest control occupations
\item \code{456} supervisors, personal service occupations
\item \code{457} barbers
\item \code{458} hairdressers and cosmetologists
\item \code{459} attendants, amusement and recreation facilities
\item \code{463} guides
\item \code{464} ushers
\item \code{465} public transportation attendants
\item \code{466} baggage porters and bellhops
\item \code{467} welfare service aides
\item \code{468} child care workers, except private household
\item \code{469} personal service occupations, n.e.c.
\item \code{473} farmers, except horticultural
\item \code{474} horticultural specialty farmers
\item \code{475} managers, farms, except horticultural
\item \code{476} managers, horticultural specialty farms
\item \code{477} supervisors, farm workers
\item \code{479} farm workers
\item \code{483} marine life cultivation workers
\item \code{484} nursery workers
\item \code{485} supervisors, related agricultural occupations
\item \code{486} groundskeepers and gardeners, except farm
\item \code{487} animal caretakers, except farm
\item \code{488} graders and sorters, agricultural products
\item \code{489} inspectors, agricultural products
\item \code{494} supervisors, forestry and logging workers
\item \code{495} forestry workers, except logging
\item \code{496} timber cutting and logging occupations
\item \code{497} captains and other officers, fishing vessels
\item \code{498} fishers
\item \code{499} hunters and trappers
\item \code{503} supervisors, mechanics and repairers
\item \code{505} automobile mechanics, except apprentices
\item \code{506} automobile mechanic apprentices
\item \code{507} bus, truck, and stationary engine mechanics
\item \code{508} aircraft engine mechanics
\item \code{509} small engine repairers
\item \code{514} automobile body and related repairers
\item \code{515} aircraft mechanics, excluding engine
\item \code{516} heavy equipment mechanics
\item \code{517} farm equipment mechanics
\item \code{518} industrial machinery repairers
\item \code{519} machinery maintenance occupations
\item \code{523} electronic repairers, communications and industrial equipment
\item \code{525} data processing equipment repairers
\item \code{526} household appliance and power tool repairers
\item \code{527} telephone line installers and repairers
\item \code{529} telephone installers and repairers
\item \code{533} misc. electrical and electronic equipment repairers
\item \code{534} heating, air conditioning, and refrigeration mechanics
\item \code{535} camera, watch, and musical instrument repairers
\item \code{536} locksmiths and safe repairers
\item \code{538} office machine repairers
\item \code{539} mechanical controls and valve repairers
\item \code{543} elevator installers and repairers
\item \code{544} millwrights
\item \code{547} specified mechanics and repairers, n.e.c.
\item \code{549} not specified mechanics and repairers
\item \code{553} supervisors, brickmasons, stonemasons, and title setters
\item \code{554} supervisors, carpenters and related work
\item \code{555} supervisors, electricians and power transmission installers
\item \code{556} supervisors, painters, paperhangers, and plasterers
\item \code{557} supervisors, plumbers, pipefitters, and steamfitters
\item \code{558} supervisors, n.e.c.
\item \code{563} brickmasons and stonemasons, except apprentices
\item \code{564} brickmasons and stonemasons apprentices
\item \code{565} tile setters, hard and soft
\item \code{566} carpet installers
\item \code{567} carpenters, except apprentices
\item \code{569} carpenter apprentices
\item \code{573} drywall installers
\item \code{575} electricians, except apprentices
\item \code{576} electrician apprentices
\item \code{577} electrical power installers and repairers
\item \code{579} painters, construction and maintenance
\item \code{583} paperhangers
\item \code{584} plasterers
\item \code{585} plumbers, pipefitters, and steamfitters, except apprentices
\item \code{587} plumber, pipefitter, and steamfitter apprentices
\item \code{588} concrete and terrazzo finishers
\item \code{589} glaziers
\item \code{593} insulation workers
\item \code{594} paving, surfacing, and tamping equipment operators
\item \code{595} roofers
\item \code{596} sheetmetal duct installers
\item \code{597} structural metal workers
\item \code{598} drillers, earth
\item \code{599} construction trades, n.e.c.
\item \code{613} supervisors, extractive occupations
\item \code{614} drillers, oil well
\item \code{615} explosives workers
\item \code{616} mining machine operators
\item \code{617} mining occupations, n.e.c.
\item \code{633} supervisors, production occupations
\item \code{634} tool and die makers, except apprentices
\item \code{635} tool and die maker apprentices
\item \code{636} precision assemblers, metal
\item \code{637} machinists, except apprentices
\item \code{639} machinist apprentices
\item \code{643} boilermakers
\item \code{644} precision grinders, fitters, and tool sharpeners
\item \code{645} patternmakers and model makers, metal
\item \code{646} lay-out workers
\item \code{647} precious stones and metals workers
\item \code{649} engravers, metal
\item \code{653} sheet metal workers, except apprentices
\item \code{654} sheet metal worker, apprentices
\item \code{655} miscellaneous precision metal workers
\item \code{656} patternmakers and model makers, wood
\item \code{657} cabinet makers and bench carpenters
\item \code{658} furniture and wood finishers
\item \code{659} miscellaneous precision woodworkers
\item \code{666} dressmakers
\item \code{667} tailors
\item \code{668} upholsterers
\item \code{669} shoe repairers
\item \code{673} apparel and fabric patternmakers
\item \code{674} miscellaneous precision apparel and fabric workers
\item \code{675} hand molders and shapers, except jewelers
\item \code{676} patternmakers, lay-out workers, and cutters
\item \code{677} optical goods workers
\item \code{678} dental laboratory and medical appliance technicians
\item \code{679} bookbinders
\item \code{683} electrical and electronic equipment assemblers
\item \code{684} miscellaneous precision workers, n.e.c.
\item \code{686} butchers and meat cutters
\item \code{687} bakers
\item \code{688} food batchmakers
\item \code{689} inspectors, testers, and graders
\item \code{693} adjusters and calibrators
\item \code{694} water and sewage treatment plant operators
\item \code{695} power plant operators
\item \code{696} stationary engineers
\item \code{699} miscellaneous plant and system operators
\item \code{703} lathe and turning machine set-up operators
\item \code{704} lathe and turning machine operators
\item \code{705} milling and planing machine operators
\item \code{706} punching and stamping press machine operators
\item \code{707} rolling machine operators
\item \code{708} drilling and boring machine operators
\item \code{709} grinding, abrading, buffing, and polishing machine operators
\item \code{713} forging machine operators
\item \code{714} numerical control machine operators
\item \code{715} miscellaneous metal, plastic, stone, and glass working machine operators
\item \code{717} fabricating machine operators, n.e.c.
\item \code{719} molding and casting machine operators
\item \code{723} metal plating machine operators
\item \code{724} heat treating equipment operators
\item \code{725} miscellaneous metal and plastic processing machine operators
\item \code{726} wood lathe, routing and planing machine operators
\item \code{727} sawing machine operators
\item \code{728} shaping and joining machine operators
\item \code{729} nailing and tacking machine operators
\item \code{733} miscellaneous woodworking machine operators
\item \code{734} printing machine operators
\item \code{735} photoengravers and lithographers
\item \code{736} typesetters and compositors
\item \code{737} miscellaneous printing machine operators
\item \code{738} winding and twisting machine operators
\item \code{739} knitting, looping, taping, and weaving machine operators
\item \code{743} textile cutting machine operators
\item \code{744} textile sewing machine operators
\item \code{745} shoe machine operators
\item \code{747} pressing machine operators
\item \code{748} laundering and dry cleaning machine operators
\item \code{749} miscellaneous textile machine operators
\item \code{753} cementing and gluing machine operators
\item \code{754} packaging and filling machine operators
\item \code{755} extruding and forming machine operators
\item \code{756} mixing and blending machine operators
\item \code{757} separating, filtering, and clarifying machine operators
\item \code{758} compressing and compacting machine operators
\item \code{759} painting and paint spraying machine operators
\item \code{763} roasting and baking machine operators, food
\item \code{764} washing, cleaning, and pickling machine operators
\item \code{765} folding machine operators
\item \code{766} furnace, kiln, and oven operators, except food
\item \code{768} crushing and grinding machine operators
\item \code{769} slicing and cutting machine operators
\item \code{773} motion picture projectionists
\item \code{774} photographic process machine operators
\item \code{777} miscellaneous and not specified machine operators, n.e.c.
\item \code{779} machine operators, not specified
\item \code{783} welders and cutters
\item \code{784} solderers and blazers
\item \code{785} assemblers
\item \code{786} hand cutting and trimming occupations
\item \code{787} hand molding, casting, and forming occupations
\item \code{789} hand painting,coating, and decorating occupations
\item \code{793} hand engraving and printing occupations
\item \code{794} hand grinding and polishing occupations
\item \code{795} miscellaneous hand working occupations
\item \code{796} production inspectors, checkers, and examiners
\item \code{797} production testers
\item \code{798} production samplers and weighers
\item \code{799} graders and sorters, except agricultural
\item \code{803} supervisors, motor vehicle operators
\item \code{804} truck drivers, heavy
\item \code{805} truck drivers, light
\item \code{806} driver-sales workers
\item \code{808} bus drivers
\item \code{809} taxicab drivers and chauffeurs
\item \code{813} parking lot attendants
\item \code{814} motor transportation occupations, n.e.c.
\item \code{823} railroad conductors and yardmasters
\item \code{824} locomotive operating occupations
\item \code{825} railroad brake, signal, and switch operators
\item \code{826} rail vehicle operators, n.e.c.
\item \code{828} ship captains and mates, except fishing boats
\item \code{829} sailors and deckhands
\item \code{833} marine engineers
\item \code{834} bridge, lock and lighthouse tenders
\item \code{843} supervisors, material moving equipment operators
\item \code{844} operating engineers
\item \code{845} longshore equipment operators
\item \code{848} hoist and winch operators
\item \code{849} crane and tower operators
\item \code{853} excavating and loading machine operators
\item \code{855} grader, dozer, and scraper operators
\item \code{856} industrial truck and tractor equipment operators
\item \code{859} miscellaneous material moving equipment operators
\item \code{863} handlers, equipment cleaners, helpers, and laborers, n.e.c.
\item \code{864} helpers, mechanics and repairers
\item \code{865} helpers, construction trades
\item \code{866} helpers, surveyor
\item \code{867} helpers, extractive occupations
\item \code{869} construction laborers
\item \code{873} production helpers
\item \code{875} garbage collectors
\item \code{876} stevedores
\item \code{877} stock handlers and baggers
\item \code{878} machine feeders and offbearers
\item \code{883} freight, stock, and material handlers, n.e.c.
\item \code{885} garage and service station related occupations
\item \code{887} vehicle washers and equipment cleaners
\item \code{888} hand packers and packagers
\item \code{889} laborers, except construction
\item \code{998} dk
\item \code{999} na
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abhave3}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{ablegal}},
\code{\link{abnomore}},
\code{\link{aborct}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{abspno}},
\code{\link{adoptus}},
\code{\link{adults}},
\code{\link{affrmact}},
\code{\link{age}},
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{agekdbrn}},
\code{\link{agewed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{antiwar}},
\code{\link{arrest}},
\code{\link{attend}},
\code{\link{babies}},
\code{\link{babies_exp}},
\code{\link{barate}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blksimp}},
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{bored}},
\code{\link{born}},
\code{\link{brass}},
\code{\link{broadband}},
\code{\link{burglr}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{capfirm}},
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{cathtemp}},
\code{\link{childs}},
\code{\link{childs_exp}},
\code{\link{childsinhh}},
\code{\link{chlddth}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldsoon}},
\code{\link{cigweek}},
\code{\link{civic}},
\code{\link{civright}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{cobarate}},
\code{\link{codeg}},
\code{\link{coden}},
\code{\link{codipged}},
\code{\link{coeduc}},
\code{\link{coeftotlt}},
\code{\link{coevwork}},
\code{\link{cofund}},
\code{\link{cogradtounder}},
\code{\link{cohrs1}},
\code{\link{cohrs2}},
\code{\link{coind10}},
\code{\link{cojail}},
\code{\link{cojew}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{color}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{commute}},
\code{\link{comprend}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlabory}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{contemp}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{convictd}},
\code{\link{coocc10}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{coother}},
\code{\link{copres10}},
\code{\link{copunish}},
\code{\link{corel}},
\code{\link{cosector}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{defwrkev}},
\code{\link{defwrknw}},
\code{\link{degree}},
\code{\link{dejavu}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{dipged}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{dotdata}},
\code{\link{dotged}},
\code{\link{dotpeop}},
\code{\link{dotpres}},
\code{\link{dotsvp}},
\code{\link{dotthng}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{earnrs}},
\code{\link{educ}},
\code{\link{educop}},
\code{\link{eftotlt}},
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{era}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{esp}},
\code{\link{evsmoke}},
\code{\link{evwork}},
\code{\link{fair}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{family16sex}},
\code{\link{father}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{febear}},
\code{\link{febrass}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fedirty}},
\code{\link{fefam}},
\code{\link{fefight}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{fehelp}},
\code{\link{fehlpmil}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{femilop}},
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{fenumok}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrite}},
\code{\link{fightair}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{form}},
\code{\link{found}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govcare}},
\code{\link{grace}},
\code{\link{gradtounder}},
\code{\link{granborn}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{guerilla}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}},
\code{\link{gunner}},
\code{\link{gunnum}},
\code{\link{hapcohab}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{healer}},
\code{\link{health}},
\code{\link{helpblk}},
\code{\link{helpful}},
\code{\link{helpnot}},
\code{\link{helpoth}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitnum}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{hompop}},
\code{\link{hompop_exp}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{id}()},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{immcrime}},
\code{\link{immjobs}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imppol}},
\code{\link{imprelax}},
\code{\link{impwork}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus07}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewtemp}},
\code{\link{jobaccmp}},
\code{\link{jobfind}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{jobtrain}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kidssol}},
\code{\link{king}},
\code{\link{landwar}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{lessnuke}},
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{liberatr}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{libtemp}},
\code{\link{life}},
\code{\link{likediff}},
\code{\link{lockedup}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madeath}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maindstatus}},
\code{\link{maindtag}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{maocc10}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{maocctag}},
\code{\link{mapa}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{mapres80}},
\code{\link{marital}},
\code{\link{marsame}},
\code{\link{marsame1}},
\code{\link{marsamey}},
\code{\link{martype}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{mawrkslf}},
\code{\link{mechanic}},
\code{\link{meserve}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{mindbody}},
\code{\link{mininc}},
\code{\link{minmilop}},
\code{\link{mobile16}},
\code{\link{mobiledata}},
\code{\link{morenuke}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{neargod}},
\code{\link{news}()},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{nurse}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{oth16}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{owngun}},
\code{\link{paattend}},
\code{\link{padeath}},
\code{\link{padeg}},
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{paeduc}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{palefull}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres10}},
\code{\link{papres105plus}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{parborn}},
\code{\link{parsol}},
\code{\link{partfull}},
\code{\link{partyid}},
\code{\link{pawrkgrw}},
\code{\link{pawrkslf}},
\code{\link{phone}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{pistol}},
\code{\link{polabuse}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{polhitok}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{polmurdr}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{popespks}},
\code{\link{popular}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{premarsx}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prestg10}},
\code{\link{prestg105plus}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{preteen}},
\code{\link{preteen_exp}},
\code{\link{privacy}},
\code{\link{prottemp}},
\code{\link{prowar}},
\code{\link{quitsmk}},
\code{\link{racavoid}},
\code{\link{raccare}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdif5}},
\code{\link{racdif5x}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs15}},
\code{\link{raceacs2}},
\code{\link{raceacs4}},
\code{\link{raceacs6}},
\code{\link{raceacs8}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{racjoin}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{racthink}},
\code{\link{ractrust}},
\code{\link{racwrite}},
\code{\link{radiohrs}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{reg16}},
\code{\link{region}},
\code{\link{relgrade}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{res16}},
\code{\link{resdefwk}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{restact}},
\code{\link{richwork}},
\code{\link{rifle}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{robbry}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{rowngun}},
\code{\link{rushed}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfin}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{satjob}},
\code{\link{school}},
\code{\link{sector}},
\code{\link{sex}},
\code{\link{sexeduc}},
\code{\link{sexnow2}},
\code{\link{shotgun}},
\code{\link{sibdeath}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{size}},
\code{\link{smoke}},
\code{\link{smokecig}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{spanking}},
\code{\link{spattend}},
\code{\link{spbarate}},
\code{\link{spdeath}},
\code{\link{spdeg}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spdipged}},
\code{\link{spdotdat}},
\code{\link{spdotged}},
\code{\link{spdotpeo}},
\code{\link{spdotpre}},
\code{\link{spdotsvp}},
\code{\link{spdotthn}},
\code{\link{spdue}},
\code{\link{speduc}},
\code{\link{speftotlt}},
\code{\link{spevwork}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spgradtounder}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spirits}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sppaid}},
\code{\link{sppres}},
\code{\link{sppres10}},
\code{\link{sppres105plus}},
\code{\link{sppres80}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{spsector}},
\code{\link{spwrkslf}},
\code{\link{spwrksta}},
\code{\link{srcbelt}},
\code{\link{stresswk}},
\code{\link{strike}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{teenpill}},
\code{\link{teens}},
\code{\link{teens_exp}},
\code{\link{teensex}},
\code{\link{thnkself}},
\code{\link{ticket}},
\code{\link{tithing}},
\code{\link{transair}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{tvhours}},
\code{\link{typist}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrelat}},
\code{\link{upgrade}()},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{vetaid}},
\code{\link{vetfam}},
\code{\link{vetfamnw}},
\code{\link{visions}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{weekswrk}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{widowed}},
\code{\link{wirtap}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}},
\code{\link{workhard}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{wrkwayup}},
\code{\link{xmarsex}},
\code{\link{xmovie}},
\code{\link{xmovie1}},
\code{\link{xnorcsiz}},
\code{\link{year}()},
\code{\link{youngen}}
}
\concept{Core}
\keyword{variable}
