% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{maocctag}
\alias{maocctag}
\title{Maocc10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
maocctag
}
\details{
Question MAOCC10 Coding Tag
}
\section{Values}{

\itemize{
\item \verb{[1]} found
\item \verb{[2]} not found, assigned
\item \verb{[3]} not found, not assigned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5998/vshow}{GSS Data Explorer page for \code{maocctag}}.

Counts by year:\tabular{llllll}{
   year \tab found \tab not found, assigned \tab not found, not assigned \tab iap \tab Total \cr
   1972 \tab 702 \tab 911 \tab - \tab - \tab 1613 \cr
   1973 \tab 1234 \tab 270 \tab - \tab - \tab 1504 \cr
   1974 \tab 1217 \tab 267 \tab - \tab - \tab 1484 \cr
   1975 \tab 976 \tab 514 \tab - \tab - \tab 1490 \cr
   1976 \tab 1230 \tab 269 \tab - \tab - \tab 1499 \cr
   1977 \tab 817 \tab 713 \tab - \tab - \tab 1530 \cr
   1978 \tab 1518 \tab 14 \tab - \tab - \tab 1532 \cr
   1980 \tab 1293 \tab 175 \tab - \tab - \tab 1468 \cr
   1982 \tab 1696 \tab 164 \tab - \tab - \tab 1860 \cr
   1983 \tab 1475 \tab 124 \tab - \tab - \tab 1599 \cr
   1984 \tab 1458 \tab 15 \tab - \tab - \tab 1473 \cr
   1985 \tab 1296 \tab 238 \tab - \tab - \tab 1534 \cr
   1986 \tab 1425 \tab 45 \tab - \tab - \tab 1470 \cr
   1987 \tab 1511 \tab 308 \tab - \tab - \tab 1819 \cr
   1988 \tab 1460 \tab 21 \tab - \tab - \tab 1481 \cr
   1989 \tab 1420 \tab 117 \tab - \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 14 \tab - \tab - \tab 1372 \cr
   1991 \tab 1499 \tab 18 \tab - \tab - \tab 1517 \cr
   1993 \tab 1545 \tab 61 \tab - \tab - \tab 1606 \cr
   1994 \tab 2938 \tab 50 \tab 4 \tab - \tab 2992 \cr
   1996 \tab 2883 \tab 20 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 2801 \tab 28 \tab 3 \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   Total \tab 50723 \tab 4356 \tab 8 \tab 0 \tab 55087 \cr
}
}

\seealso{
Other Core: 
\code{\link{childs}},
\code{\link{maind10}},
\code{\link{maindstatus}},
\code{\link{maindtag}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{sibs}}
}
\concept{Core}
\keyword{variable}
