% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{mapaid}
\alias{mapaid}
\title{Should working women have paid maternity leave}
\source{
General Social Survey https://gss.norc.org
}
\description{
mapaid
}
\details{
Question Do you agree or disagree . . .
A. Working women should receive paid maternity leave when they have a baby.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4412/vshow}{GSS Data Explorer page for \code{mapaid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 384 \tab 633 \tab 148 \tab 150 \tab 26 \tab 26 \tab 80 \tab 2992 \cr
   2002 \tab 1594 \tab 571 \tab 411 \tab 98 \tab 56 \tab 19 \tab 15 \tab 1 \tab 2765 \cr
   Total \tab 3139 \tab 955 \tab 1044 \tab 246 \tab 206 \tab 45 \tab 41 \tab 81 \tab 5757 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
