% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{mapres80}
\alias{mapres80}
\title{Mothers occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mapres80
}
\details{
Question A. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up?  That is, what was her job called?
OCCUPATION: (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What did she actually do in that job? Tell me, what were some of her main duties?
C. What kind of place did she work for?
INDUSTRY: (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was she self-employed, or did she work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Core: 
\code{\link{maind80}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mawrkslf}},
\code{\link{paind10}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{papres10}},
\code{\link{papres105plus}}

Other Occupation: 
\code{\link{maind80}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mawrkslf}},
\code{\link{paind10}}
}
\concept{Core}
\concept{Occupation}
\concept{Occupational Prestige}
\keyword{variable}
