% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_588.R
\name{mar11}
\alias{mar11}
\title{Marital status of 11th person (visitor)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mar11
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
4. Marital status of eleventh person
}
\section{Values}{

\itemize{
\item \code{1} married
\item \code{2} widowed
\item \code{3} divorced
\item \code{4} separated
\item \code{5} never married
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5395/vshow}{GSS Data Explorer page for \code{mar11}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab married \tab widowed \tab divorced \tab separated \tab never married \tab no answer \tab don't know \tab not available in this release \tab Total \cr
   1975 \tab 1459 \tab 9 \tab 7 \tab 3 \tab 3 \tab 9 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1494 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1513 \tab 4 \tab 3 \tab 3 \tab - \tab 5 \tab 2 \tab - \tab - \tab 1530 \cr
   1978 \tab 1513 \tab 4 \tab 1 \tab 2 \tab 2 \tab 8 \tab 2 \tab - \tab - \tab 1532 \cr
   1980 \tab 1441 \tab 5 \tab 4 \tab 5 \tab 5 \tab 7 \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 1845 \tab 3 \tab - \tab 2 \tab 3 \tab 4 \tab 3 \tab - \tab - \tab 1860 \cr
   1983 \tab 1576 \tab 2 \tab 7 \tab - \tab 2 \tab 9 \tab 3 \tab - \tab - \tab 1599 \cr
   1984 \tab 1443 \tab 7 \tab 2 \tab 3 \tab 4 \tab 9 \tab 5 \tab - \tab - \tab 1473 \cr
   1985 \tab 1501 \tab 6 \tab 4 \tab 3 \tab 2 \tab 5 \tab 13 \tab - \tab - \tab 1534 \cr
   1986 \tab 1454 \tab 3 \tab 2 \tab 1 \tab - \tab 9 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 1797 \tab 8 \tab 3 \tab 3 \tab 2 \tab 5 \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 1470 \tab 4 \tab 1 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1522 \tab 5 \tab 2 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1359 \tab 2 \tab 4 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1498 \tab 4 \tab 1 \tab 1 \tab 1 \tab 12 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1598 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2972 \tab 6 \tab 1 \tab 3 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab 2992 \cr
   1996 \tab 2884 \tab 5 \tab - \tab - \tab 4 \tab 7 \tab 4 \tab - \tab - \tab 2904 \cr
   1998 \tab 2810 \tab 11 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 2810 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 2759 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2798 \tab 1 \tab 1 \tab 3 \tab - \tab 8 \tab - \tab 1 \tab - \tab 2812 \cr
   2006 \tab 4463 \tab 11 \tab 3 \tab 4 \tab 1 \tab 28 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2007 \tab 2 \tab 1 \tab 4 \tab 1 \tab 8 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2019 \tab 6 \tab 1 \tab 2 \tab - \tab 14 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1955 \tab 5 \tab 1 \tab 5 \tab 1 \tab 7 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2522 \tab 2 \tab - \tab 3 \tab 2 \tab 7 \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2841 \tab 4 \tab 3 \tab 5 \tab 3 \tab 9 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2317 \tab 2 \tab 3 \tab 13 \tab 1 \tab 8 \tab 1 \tab 3 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3510 \tab 6 \tab 1 \tab 7 \tab 3 \tab 17 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 63150 \tab 132 \tab 61 \tab 86 \tab 49 \tab 227 \tab 45 \tab 7 \tab 4032 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away11}},
\code{\link{away12}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{mar12}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{relate12}},
\code{\link{where11}}
}
\concept{Household Members and Structure}
\keyword{variable}
