% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{marasian}
\alias{marasian}
\title{Close relative marry asian}
\source{
General Social Survey https://gss.norc.org
}
\description{
marasian
}
\details{
Question What about having a close relative marry a Jewish person?  Would you be very in favor of it happening, somewhat In favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
C. An Asian American person?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1223/vshow}{GSS Data Explorer page for \code{marasian}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 39 \tab 90 \tab 632 \tab 373 \tab 205 \tab 24 \tab 9 \tab - \tab 1372 \cr
   2000 \tab 1419 \tab 202 \tab 236 \tab 609 \tab 159 \tab 115 \tab 62 \tab 15 \tab - \tab 2817 \cr
   2004 \tab 1906 \tab 106 \tab 155 \tab 482 \tab 99 \tab 58 \tab 3 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 303 \tab 335 \tab 1000 \tab 215 \tab 125 \tab 11 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 229 \tab 213 \tab 673 \tab 132 \tab 70 \tab 9 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 263 \tab 234 \tab 743 \tab 123 \tab 58 \tab 5 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 245 \tab 196 \tab 701 \tab 99 \tab 47 \tab 12 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 329 \tab 272 \tab 895 \tab 103 \tab 56 \tab 14 \tab 6 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 396 \tab 291 \tab 1021 \tab 118 \tab 50 \tab 7 \tab 5 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 367 \tab 258 \tab 793 \tab 87 \tab 41 \tab 7 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 1318 \tab 513 \tab 242 \tab 1816 \tab 76 \tab 42 \tab - \tab 2 \tab 23 \tab 4032 \cr
   2022 \tab 1168 \tab 604 \tab 274 \tab 1354 \tab 79 \tab 48 \tab 12 \tab 4 \tab 1 \tab 3544 \cr
   Total \tab 12940 \tab 3596 \tab 2796 \tab 10719 \tab 1663 \tab 915 \tab 166 \tab 62 \tab 24 \tab 32881 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Asian Americans: 
\code{\link{acqasian}},
\code{\link{acqfmasn}},
\code{\link{acqnhasn}},
\code{\link{acqvaasn}},
\code{\link{acqwkasn}},
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asngovt}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{comasn}},
\code{\link{fairasns}},
\code{\link{famasns}},
\code{\link{fareasns}},
\code{\link{feelasns}},
\code{\link{hspasn10}},
\code{\link{influasn}},
\code{\link{intlasns}},
\code{\link{knwasn}},
\code{\link{liveasns}},
\code{\link{patrasns}},
\code{\link{trtasian}},
\code{\link{usasn}},
\code{\link{violasns}},
\code{\link{wlthasns}},
\code{\link{workasns}}
}
\concept{Asian Americans}
\concept{Intergroup Relations}
\keyword{variable}
