% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{marblk}
\alias{marblk}
\title{Close relative marry black person}
\source{
General Social Survey https://gss.norc.org
}
\description{
marblk
}
\details{
Question What about having a close relative marry a Black person? Would you be very in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1996, 1998, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1222/vshow}{GSS Data Explorer page for \code{marblk}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 96 \tab 61 \tab 407 \tab 342 \tab 441 \tab 15 \tab 10 \tab - \tab 1372 \cr
   1996 \tab 1935 \tab 125 \tab 108 \tab 373 \tab 152 \tab 192 \tab 10 \tab 9 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 251 \tab 198 \tab 738 \tab 313 \tab 345 \tab 15 \tab 11 \tab - \tab 2832 \cr
   2000 \tab 486 \tab 413 \tab 278 \tab 875 \tab 307 \tab 402 \tab 30 \tab 26 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 186 \tab 102 \tab 356 \tab 127 \tab 131 \tab 6 \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 131 \tab 106 \tab 432 \tab 124 \tab 108 \tab 1 \tab 4 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 333 \tab 263 \tab 900 \tab 235 \tab 248 \tab 10 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 235 \tab 189 \tab 612 \tab 138 \tab 145 \tab 5 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 293 \tab 189 \tab 680 \tab 136 \tab 123 \tab 4 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 241 \tab 169 \tab 665 \tab 112 \tab 104 \tab 9 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 343 \tab 240 \tab 844 \tab 127 \tab 104 \tab 9 \tab 8 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 405 \tab 254 \tab 982 \tab 140 \tab 99 \tab 4 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 386 \tab 223 \tab 757 \tab 108 \tab 76 \tab 3 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 1318 \tab 546 \tab 196 \tab 1747 \tab 144 \tab 65 \tab - \tab 1 \tab 15 \tab 4032 \cr
   2022 \tab 1168 \tab 640 \tab 236 \tab 1327 \tab 97 \tab 61 \tab 9 \tab 4 \tab 2 \tab 3544 \cr
   2024 \tab 1126 \tab 581 \tab 200 \tab 1225 \tab 99 \tab 52 \tab 13 \tab 4 \tab 9 \tab 3309 \cr
   Total \tab 17886 \tab 5205 \tab 3012 \tab 12920 \tab 2701 \tab 2696 \tab 143 \tab 102 \tab 26 \tab 44691 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Blacks: 
\code{\link{acqblack}},
\code{\link{acqfmblk}},
\code{\link{acqnhblk}},
\code{\link{acqvablk}},
\code{\link{acqwkblk}},
\code{\link{admirblk}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkgovt}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blksimp}},
\code{\link{blkwrk}},
\code{\link{closeblk}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{comblk}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fairblks}},
\code{\link{famblks}},
\code{\link{fareblks}},
\code{\link{feelblks}},
\code{\link{influblk}},
\code{\link{intlblks}},
\code{\link{jobaff}},
\code{\link{knwblk}},
\code{\link{liveblks}},
\code{\link{patrblks}},
\code{\link{symptblk}},
\code{\link{trtblack}},
\code{\link{usblk}},
\code{\link{violblks}},
\code{\link{wlthblks}},
\code{\link{workblks}},
\code{\link{wrkwayup}}
}
\concept{Blacks}
\concept{Intergroup Relations}
\keyword{variable}
