% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{marcohab}
\alias{marcohab}
\title{Cohabitation status}
\source{
General Social Survey https://gss.norc.org
}
\description{
marcohab
}
\details{
Question Cohabitating status
}
\section{Values}{

\itemize{
\item \verb{[1]} married
\item \verb{[2]} not married, cohabitating partner
\item \verb{[3]} not married, no cohabitating partner
\item \verb{[4]} not married, missing on cohabitating
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6320/vshow}{GSS Data Explorer page for \code{marcohab}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab married \tab not married, cohabitating partner \tab not married, no cohabitating partner \tab not married, missing on cohabitating \tab no answer \tab Total \cr
   2012 \tab - \tab 900 \tab 168 \tab 896 \tab 10 \tab - \tab 1974 \cr
   2014 \tab - \tab 1158 \tab 210 \tab 1161 \tab 6 \tab 3 \tab 2538 \cr
   2016 \tab 16 \tab 1212 \tab 263 \tab 1376 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 998 \tab 231 \tab 1111 \tab 7 \tab 1 \tab 2348 \cr
   2021 \tab - \tab 1999 \tab 372 \tab 1648 \tab 12 \tab 1 \tab 4032 \cr
   2022 \tab - \tab 1470 \tab 310 \tab 1725 \tab 31 \tab 8 \tab 3544 \cr
   2024 \tab - \tab 1363 \tab 261 \tab 1652 \tab 25 \tab 8 \tab 3309 \cr
   Total \tab 16 \tab 9100 \tab 1815 \tab 9569 \tab 91 \tab 21 \tab 20612 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{bestfrd}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{comyear}},
\code{\link{cousins}},
\code{\link{cowrkfrd}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{hapunhap}},
\code{\link{helpfrds}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{inlawnum}},
\code{\link{kidcall}},
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidpars}},
\code{\link{kidvisit}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{localgvt}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{malive}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{neifrd}},
\code{\link{niecenep}},
\code{\link{othfrd}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{palive}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{peocntct}},
\code{\link{posslq}},
\code{\link{posslqy}},
\code{\link{relcall}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{talkedto}},
\code{\link{trustpeo}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
