% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{mardiv}
\alias{mardiv}
\title{Marriage not taken seriously when div is easy}
\source{
General Social Survey https://gss.norc.org
}
\description{
mardiv
}
\details{
Question Do you agree or disagree?
I. Couples don''t take marriage seriously enough when divorce is easily available.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4372/vshow}{GSS Data Explorer page for \code{mardiv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 293 \tab 603 \tab 227 \tab 195 \tab 47 \tab 26 \tab 23 \tab 1481 \cr
   Total \tab 67 \tab 293 \tab 603 \tab 227 \tab 195 \tab 47 \tab 26 \tab 23 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{marhomo}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{threkids}},
\code{\link{twokids}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
