% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{marelig}
\alias{marelig}
\title{What was mother's religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
marelig
}
\details{
Question What was your mother''s religious preference when you were growing up?  Was it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1058/vshow}{GSS Data Explorer page for \code{marelig}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab no answer \tab don't know \tab Total \cr
   1988 \tab 48 \tab 910 \tab 400 \tab 32 \tab 50 \tab 36 \tab 5 \tab - \tab 1481 \cr
   2008 \tab 1969 \tab 10 \tab - \tab 38 \tab 5 \tab - \tab - \tab 1 \tab 2023 \cr
   2012 \tab 1931 \tab 6 \tab 5 \tab 28 \tab 3 \tab 1 \tab - \tab - \tab 1974 \cr
   Total \tab 3948 \tab 926 \tab 405 \tab 98 \tab 58 \tab 37 \tab 5 \tab 1 \tab 5478 \cr
}
}

\seealso{
Other Religion: 
\code{\link{maden}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}}
}
\concept{Religion}
\keyword{variable}
