% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marfin}
\alias{marfin}
\title{Financial security main advtg of marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
marfin
}
\details{
Question Do you agree or disagree?
C. The main advantage of marriage is that it gives financial security.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4366/vshow}{GSS Data Explorer page for \code{marfin}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 52 \tab 172 \tab 250 \tab 672 \tab 218 \tab 28 \tab 22 \tab 1481 \cr
   1994 \tab 1545 \tab 48 \tab 193 \tab 305 \tab 687 \tab 173 \tab 18 \tab 23 \tab 2992 \cr
   Total \tab 1612 \tab 100 \tab 365 \tab 555 \tab 1359 \tab 391 \tab 46 \tab 45 \tab 4473 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
