% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marfree}
\alias{marfree}
\title{Personal freedom more impt than marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
marfree
}
\details{
Question Do you agree or disagree?
B. Personal freedom is more important than the companionship of marriage.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4365/vshow}{GSS Data Explorer page for \code{marfree}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 54 \tab 160 \tab 345 \tab 633 \tab 144 \tab 50 \tab 28 \tab 1481 \cr
   1996 \tab 1444 \tab 61 \tab 152 \tab 218 \tab 696 \tab 308 \tab 23 \tab 2 \tab 2904 \cr
   Total \tab 1511 \tab 115 \tab 312 \tab 563 \tab 1329 \tab 452 \tab 73 \tab 30 \tab 4385 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{Gender}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
