% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marhappy}
\alias{marhappy}
\title{Married people happier than unmarried}
\source{
General Social Survey https://gss.norc.org
}
\description{
marhappy
}
\details{
Question Do you agree or disagree?
A. Married people are generally happier than unmarried people.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4364/vshow}{GSS Data Explorer page for \code{marhappy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 233 \tab 485 \tab 411 \tab 184 \tab 30 \tab 53 \tab 18 \tab 1481 \cr
   1994 \tab 1545 \tab 149 \tab 496 \tab 439 \tab 246 \tab 57 \tab 39 \tab 21 \tab 2992 \cr
   2002 \tab 1594 \tab 191 \tab 272 \tab 367 \tab 196 \tab 103 \tab 41 \tab 1 \tab 2765 \cr
   2012 \tab 672 \tab 100 \tab 348 \tab 383 \tab 327 \tab 59 \tab 76 \tab 9 \tab 1974 \cr
   Total \tab 3878 \tab 673 \tab 1601 \tab 1600 \tab 953 \tab 249 \tab 209 \tab 49 \tab 9212 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
