% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{marhisp}
\alias{marhisp}
\title{Close relative marry hispanic}
\source{
General Social Survey https://gss.norc.org
}
\description{
marhisp
}
\details{
Question What about having a close relative marry a Hispanic American person? Would you be very in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1224/vshow}{GSS Data Explorer page for \code{marhisp}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 62 \tab 94 \tab 636 \tab 343 \tab 208 \tab 19 \tab 10 \tab - \tab 1372 \cr
   2000 \tab 1419 \tab 243 \tab 228 \tab 597 \tab 158 \tab 111 \tab 47 \tab 14 \tab - \tab 2817 \cr
   2004 \tab 1906 \tab 121 \tab 140 \tab 484 \tab 100 \tab 54 \tab 3 \tab 4 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 360 \tab 321 \tab 950 \tab 227 \tab 123 \tab 8 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 250 \tab 209 \tab 651 \tab 129 \tab 80 \tab 6 \tab 4 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 267 \tab 230 \tab 734 \tab 124 \tab 65 \tab 5 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 260 \tab 198 \tab 688 \tab 87 \tab 52 \tab 15 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 356 \tab 268 \tab 880 \tab 98 \tab 55 \tab 12 \tab 6 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 412 \tab 299 \tab 1013 \tab 109 \tab 42 \tab 7 \tab 6 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 381 \tab 241 \tab 788 \tab 87 \tab 48 \tab 6 \tab 8 \tab - \tab 2348 \cr
   2021 \tab 1318 \tab 517 \tab 257 \tab 1794 \tab 82 \tab 40 \tab - \tab 2 \tab 22 \tab 4032 \cr
   2022 \tab 1168 \tab 630 \tab 264 \tab 1356 \tab 69 \tab 39 \tab 9 \tab 4 \tab 5 \tab 3544 \cr
   Total \tab 12940 \tab 3859 \tab 2749 \tab 10571 \tab 1613 \tab 917 \tab 137 \tab 68 \tab 27 \tab 32881 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{lessfare}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{workfare}}
}
\concept{Intergroup Relations}
\keyword{variable}
