% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{marhomo}
\alias{marhomo}
\title{Homosexuals should have right to marry}
\source{
General Social Survey https://gss.norc.org
}
\description{
marhomo
}
\details{
Question Do you agree or disagree?
J. Homosexual couples should have the right to marry one another.
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2004 \cr
   A/C/- \tab 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4373/vshow}{GSS Data Explorer page for \code{marhomo}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 44 \tab 118 \tab 196 \tab 336 \tab 613 \tab 82 \tab 25 \tab - \tab 1481 \cr
   2004 \tab 1596 \tab 135 \tab 216 \tab 173 \tab 240 \tab 420 \tab 22 \tab 10 \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1358 \tab 1265 \tab 500 \tab 419 \tab 179 \tab 295 \tab 5 \tab 1 \tab 10 \tab 4032 \cr
   2022 \tab 1203 \tab 1047 \tab 499 \tab 310 \tab 189 \tab 262 \tab 14 \tab 13 \tab 7 \tab 3544 \cr
   Total \tab 22528 \tab 2491 \tab 1333 \tab 1098 \tab 944 \tab 1590 \tab 123 \tab 49 \tab 17 \tab 30173 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{mardiv}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{threkids}},
\code{\link{twokids}}

Other Core: 
\code{\link{marhomo1}}
}
\concept{Core}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
