% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{marhomo1}
\alias{marhomo1}
\title{Homosexuals have right to marry}
\source{
General Social Survey https://gss.norc.org
}
\description{
marhomo1
}
\details{
Question Do you agree or disagree with the following statement?
Homosexual couples have the right to marry one another.
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2006, 2008, 2010, 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7428/vshow}{GSS Data Explorer page for \code{marhomo1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2507 \tab - \tab 307 \tab 391 \tab 260 \tab 329 \tab 695 \tab 17 \tab 4 \tab 4510 \cr
   2008 \tab 671 \tab - \tab 223 \tab 307 \tab 170 \tab 213 \tab 423 \tab 14 \tab 2 \tab 2023 \cr
   2010 \tab 763 \tab - \tab 270 \tab 316 \tab 167 \tab 191 \tab 318 \tab 13 \tab 6 \tab 2044 \cr
   2012 \tab 666 \tab - \tab 323 \tab 315 \tab 152 \tab 183 \tab 314 \tab 14 \tab 7 \tab 1974 \cr
   2014 \tab 824 \tab - \tab 532 \tab 423 \tab 185 \tab 223 \tab 327 \tab 12 \tab 12 \tab 2538 \cr
   2016 \tab 977 \tab - \tab 640 \tab 464 \tab 224 \tab 223 \tab 311 \tab 16 \tab 12 \tab 2867 \cr
   2018 \tab 774 \tab - \tab 619 \tab 416 \tab 157 \tab 148 \tab 214 \tab 12 \tab 8 \tab 2348 \cr
   Total \tab 7182 \tab 0 \tab 2914 \tab 2632 \tab 1315 \tab 1510 \tab 2602 \tab 98 \tab 51 \tab 18304 \cr
}
}

\seealso{
Other Core: 
\code{\link{marhomo}}
}
\concept{Core}
\keyword{variable}
