% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marlegit}
\alias{marlegit}
\title{Those wanting kids should get married}
\source{
General Social Survey https://gss.norc.org
}
\description{
marlegit
}
\details{
Question Do you agree or disagree?
F. People who want children ought to get married.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4369/vshow}{GSS Data Explorer page for \code{marlegit}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 484 \tab 529 \tab 165 \tab 156 \tab 48 \tab 8 \tab 24 \tab - \tab 1481 \cr
   1994 \tab 1545 \tab 434 \tab 597 \tab 145 \tab 175 \tab 55 \tab 15 \tab 26 \tab - \tab 2992 \cr
   2002 \tab 1594 \tab 450 \tab 302 \tab 182 \tab 141 \tab 77 \tab 17 \tab 2 \tab - \tab 2765 \cr
   2012 \tab 672 \tab 244 \tab 544 \tab 227 \tab 221 \tab 36 \tab 24 \tab 6 \tab - \tab 1974 \cr
   2022 \tab 2381 \tab 147 \tab 317 \tab 325 \tab 233 \tab 88 \tab 41 \tab 5 \tab 7 \tab 3544 \cr
   Total \tab 6259 \tab 1759 \tab 2289 \tab 1044 \tab 926 \tab 304 \tab 105 \tab 63 \tab 7 \tab 12756 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
