% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marmakid}
\alias{marmakid}
\title{Single ma can raise kids well as couple}
\source{
General Social Survey https://gss.norc.org
}
\description{
marmakid
}
\details{
Question Do you agree or disagree?
G. A single mother can bring up her child as well as a married couple.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4370/vshow}{GSS Data Explorer page for \code{marmakid}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 111 \tab 408 \tab 223 \tab 477 \tab 156 \tab 16 \tab 23 \tab - \tab 1481 \cr
   2022 \tab 2381 \tab 260 \tab 466 \tab 158 \tab 195 \tab 46 \tab 29 \tab 4 \tab 5 \tab 3544 \cr
   Total \tab 2448 \tab 371 \tab 874 \tab 381 \tab 672 \tab 202 \tab 45 \tab 27 \tab 5 \tab 5025 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
