% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marnomar}
\alias{marnomar}
\title{Bad marriage better than none at all}
\source{
General Social Survey https://gss.norc.org
}
\description{
marnomar
}
\details{
Question Do you agree or disagree?
E. It is better to have a bad marriage than no marriage at all.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4368/vshow}{GSS Data Explorer page for \code{marnomar}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 19 \tab 44 \tab 81 \tab 583 \tab 641 \tab 23 \tab 23 \tab 1481 \cr
   1994 \tab 1545 \tab 11 \tab 33 \tab 68 \tab 593 \tab 699 \tab 18 \tab 25 \tab 2992 \cr
   2002 \tab 1594 \tab 29 \tab 38 \tab 68 \tab 342 \tab 682 \tab 10 \tab 2 \tab 2765 \cr
   Total \tab 3206 \tab 59 \tab 115 \tab 217 \tab 1518 \tab 2022 \tab 51 \tab 50 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
