% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marpakid}
\alias{marpakid}
\title{Single pa can raise kids well as couple}
\source{
General Social Survey https://gss.norc.org
}
\description{
marpakid
}
\details{
Question Do you agree or disagree?
H. A single father can bring up his child as well as a married couple.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4371/vshow}{GSS Data Explorer page for \code{marpakid}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 94 \tab 348 \tab 235 \tab 529 \tab 164 \tab 18 \tab 26 \tab - \tab 1481 \cr
   2022 \tab 2381 \tab 233 \tab 448 \tab 183 \tab 201 \tab 59 \tab 30 \tab 4 \tab 5 \tab 3544 \cr
   Total \tab 2448 \tab 327 \tab 796 \tab 418 \tab 730 \tab 223 \tab 48 \tab 30 \tab 5 \tab 5025 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
