% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{marsame}
\alias{marsame}
\title{Homosexual couples should have the right to marry one another}
\source{
General Social Survey https://gss.norc.org
}
\description{
marsame
}
\details{
Question Do you agree or disagree with the following statement: Homosexual couples should have the right to marry one another.

MARSAME is a rename from MARHOMO. MARSAME is asked in sequence with PILLOK and SPANKING. Hand Card A13 contains punches 1-5.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2004 \cr
   A/C/- \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8056/vshow}{GSS Data Explorer page for \code{marsame}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 44 \tab 118 \tab 196 \tab 336 \tab 613 \tab 82 \tab 25 \tab - \tab 1481 \cr
   2004 \tab 1596 \tab 135 \tab 216 \tab 173 \tab 240 \tab 420 \tab 22 \tab 10 \tab - \tab 2812 \cr
   2021 \tab 1358 \tab 1265 \tab 500 \tab 419 \tab 179 \tab 295 \tab 5 \tab 1 \tab 10 \tab 4032 \cr
   2022 \tab 1203 \tab 1047 \tab 499 \tab 310 \tab 189 \tab 262 \tab 14 \tab 13 \tab 7 \tab 3544 \cr
   2024 \tab 2163 \tab 460 \tab 262 \tab 175 \tab 83 \tab 147 \tab 9 \tab 4 \tab 6 \tab 3309 \cr
   Total \tab 6387 \tab 2951 \tab 1595 \tab 1273 \tab 1027 \tab 1737 \tab 132 \tab 53 \tab 23 \tab 15178 \cr
}
}

\seealso{
Other Core: 
\code{\link{marsame1}}
}
\concept{Core}
\keyword{variable}
