% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{marsame1}
\alias{marsame1}
\title{Homosexual couples have right to marry}
\source{
General Social Survey https://gss.norc.org
}
\description{
marsame1
}
\details{
Question Do you agree or disagree with the following statement: Homosexual couples have the right to marry one another.

MARSAME1 is a rename from MARHOMO1.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/C \tab partial \cr
   2008 \tab A/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
   2012 \tab A/C \tab partial \cr
   2014 \tab A/C \tab partial \cr
   2016 \tab A/C \tab partial \cr
   2018 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8057/vshow}{GSS Data Explorer page for \code{marsame1}}.
}

\seealso{
Other Core: 
\code{\link{marsame}}
}
\concept{Core}
\keyword{variable}
