% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{marsamey}
\alias{marsamey}
\title{Same-sex couples should have the right to marry one another}
\source{
General Social Survey https://gss.norc.org
}
\description{
marsamey
}
\details{
Question Do you agree or disagree with the following statement: Same-sex couples  have the right to marry one another.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8058/vshow}{GSS Data Explorer page for \code{marsamey}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 404 \tab 229 \tab 163 \tab 116 \tab 158 \tab 13 \tab 2213 \tab 4 \tab 9 \tab 3309 \cr
   Total \tab 0 \tab 404 \tab 229 \tab 163 \tab 116 \tab 158 \tab 13 \tab 2213 \tab 4 \tab 9 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conlabory}},
\code{\link{norelgsp16}},
\code{\link{pawrkgrw}},
\code{\link{racdif5}}
}
\concept{Core}
\keyword{variable}
