% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{martype}
\alias{martype}
\title{Marital type}
\source{
General Social Survey https://gss.norc.org
}
\description{
martype
}
\details{
Question This based on information coded by interviewer and information in the Household Enumeration Form and other parts of the questionnaire.
}
\section{Values}{

\itemize{
\item \code{1} marriage between a man and a woman
\item \code{2} marriage between two people of the same gender
\item \code{3} not sure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/19/vshow}{GSS Data Explorer page for \code{martype}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab marriage between a man and a woman \tab marriage between two people of the same gender \tab not sure \tab Total \cr
   2012 \tab 1074 \tab 899 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 1380 \tab 1150 \tab 8 \tab - \tab 2538 \cr
   2016 \tab 1657 \tab 1195 \tab 15 \tab - \tab 2867 \cr
   2018 \tab 1351 \tab 990 \tab 7 \tab - \tab 2348 \cr
   2021 \tab 3929 \tab 83 \tab 1 \tab 19 \tab 4032 \cr
   2022 \tab 2838 \tab 682 \tab 4 \tab 20 \tab 3544 \cr
   Total \tab 12229 \tab 4999 \tab 36 \tab 39 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{agewed}},
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus10}},
\code{\link{marital}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{prestg10}},
\code{\link{prestg105plus}}
}
\concept{Core}
\keyword{variable}
