% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{marunion}
\alias{marunion}
\title{Legal marital status}
\source{
General Social Survey https://gss.norc.org
}
\description{
marunion
}
\details{
Question What is your current legal marital status?
}
\section{Values}{

\itemize{
\item \verb{[1]} legally married
\item \verb{[2]} in a civil union or registered domestic partnership
\item \verb{[3]} separated
\item \verb{[4]} not currently married or in a civil union or registered domestic partnership
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3911/vshow}{GSS Data Explorer page for \code{marunion}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab legally married \tab in a civil union or registered domestic partnership \tab separated \tab not currently married or in a civil union or registered domestic partnership \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1879 \tab 32 \tab 8 \tab 6 \tab 65 \tab 2 \tab 31 \tab 2023 \cr
   Total \tab 1879 \tab 32 \tab 8 \tab 6 \tab 65 \tab 2 \tab 31 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{toldsxor}}
}
\concept{Sexual Orientation}
\keyword{variable}
