% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{mascsee1}
\alias{mascsee1}
\title{How masculine others see r (0 to 6 scale)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mascsee1
}
\details{
Question On that same scale, how masculine do most people see you?

NOTE: FEMSELF1, FEMSEE1, MASCSELF1 and MASCSEE1 are part of a BALLOT/FORM randomization. On BALLOT=2, all items were asked at the start of the High Risk Behaviors module, and on BALLOT=3, all items were asked at the end. For FORM=1, FEMSELF1 and FEMSEE1 were asked first, and for FORM=2, MASCSELF1 and MASCSEE1 were asked first.
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[6]} very
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8189/vshow}{GSS Data Explorer page for \code{mascsee1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab not at all \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab very \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 705 \tab 177 \tab 96 \tab 141 \tab 144 \tab 269 \tab 527 \tab 1217 \tab 33 \tab 3309 \cr
   Total \tab 0 \tab 705 \tab 177 \tab 96 \tab 141 \tab 144 \tab 269 \tab 527 \tab 1217 \tab 33 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conorelgsp}},
\code{\link{family16sex}},
\code{\link{femsee1}},
\code{\link{racdif5x}},
\code{\link{sexnow2}}
}
\concept{Core}
\keyword{variable}
