% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{mascself}
\alias{mascself}
\title{How masculine r sees self}
\source{
General Social Survey https://gss.norc.org
}
\description{
mascself
}
\details{
Question How masculine do you see yourself

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[7]} very
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7961/vshow}{GSS Data Explorer page for \code{mascself}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very \tab Total \cr
   2022 \tab 2581 \tab 192 \tab 114 \tab 70 \tab 52 \tab 109 \tab 142 \tab 284 \tab 3544 \cr
   Total \tab 2581 \tab 192 \tab 114 \tab 70 \tab 52 \tab 109 \tab 142 \tab 284 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{femself}},
\code{\link{firstkidsex}},
\code{\link{nonbinkids}}
}
\concept{GSS Next}
\keyword{variable}
