% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{masei}
\alias{masei}
\title{R's mothers socioeconomic index (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
masei
}
\details{
Question Socioeconomic index of respondent's mother.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} i don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{coisco08}},
\code{\link{firstsei}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{spsei}}
}
\concept{Occupation}
\keyword{variable}
