% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_598.R
\name{masei10educ}
\alias{masei10educ}
\title{Percentage of some college educ in maocc10 based on acs 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
masei10educ
}
\details{
Question Percentage of those who had some college or more education among those who had the same occupation as in MAOCC10 in American Community Survey between 2010 and 2012 (25-64 years old only). This percentage is not from the GSS but it is used to calculate MASEI10.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} i don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5721/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 12.2 \tab 12.4 \tab 12.6 \tab 13.3 \tab 13.5 \tab 14.1 \tab 15.5 \tab 16 \tab 16.5 \tab 16.9 \tab 17 \tab 17.2 \tab 17.3 \tab 17.5 \tab 17.9 \tab 19.3 \tab 19.9 \tab 20.4 \tab 20.5 \tab 21 \tab 21.7 \tab 22 \tab 22.3 \tab 22.7 \tab 24 \tab 24.1 \tab 24.3 \tab 25.1 \tab 25.4 \tab 25.7 \tab 26 \tab 26.2 \tab 26.6 \tab 26.7 \tab 27.1 \tab 28.2 \tab 28.9 \tab 29.1 \tab 30.4 \tab 30.7 \tab 31 \tab 31.1 \tab 32.4 \tab 33.7 \tab 33.9 \tab 34.1 \tab 34.2 \tab 35 \tab 35.1 \tab 35.3 \tab 36 \tab 36.5 \tab 37 \tab 37.7 \tab 37.8 \tab 38.3 \tab 38.4 \tab 39.6 \tab 40.3 \tab 40.8 \tab 41.6 \tab 41.8 \tab 42 \tab 42.3 \tab 42.8 \tab 43.7 \tab 43.9 \tab 44 \tab 44.2 \tab 45.1 \tab 45.2 \tab 45.4 \tab 46.3 \tab 46.4 \tab 46.6 \tab 46.9 \tab 47.4 \tab 48.5 \tab 48.8 \tab 49 \tab 50 \tab 50.1 \tab 50.6 \tab 50.7 \tab 51 \tab 51.1 \tab 52.1 \tab 52.2 \tab 52.7 \tab 53 \tab 53.9 \tab 54.2 \tab 54.6 \tab 55.1 \tab 55.4 \tab 55.5 \tab 55.7 \tab 55.8 \tab 55.9 \tab 56.1 \tab 56.7 \tab 57.1 \tab 57.6 \tab 58.5 \tab 58.8 \tab 58.9 \tab 59.2 \tab 59.3 \tab 60.9 \tab 61.1 \tab 62.4 \tab 62.7 \tab 63.3 \tab 63.7 \tab 63.9 \tab 64.1 \tab 64.6 \tab 65.1 \tab 65.3 \tab 65.5 \tab 65.6 \tab 66 \tab 67.3 \tab 68.4 \tab 68.6 \tab 68.8 \tab 69.1 \tab 69.9 \tab 70.1 \tab 70.9 \tab 71.7 \tab 73 \tab 73.3 \tab 73.5 \tab 73.9 \tab 74.2 \tab 74.6 \tab 75.1 \tab 75.5 \tab 75.9 \tab 76.9 \tab 77.2 \tab 78 \tab 78.2 \tab 78.4 \tab 79.2 \tab 79.9 \tab 80.1 \tab 80.7 \tab 81.3 \tab 81.4 \tab 82.3 \tab 82.5 \tab 82.8 \tab 83 \tab 83.7 \tab 84.6 \tab 85.3 \tab 85.7 \tab 85.9 \tab 86.3 \tab 86.7 \tab 87 \tab 88.3 \tab 89.4 \tab 90.6 \tab 91.2 \tab 91.4 \tab 92 \tab 92.1 \tab 92.6 \tab 92.9 \tab 93.7 \tab 94 \tab 94.8 \tab 95.5 \tab 95.6 \tab 96.3 \tab 96.4 \tab 97.5 \tab 97.8 \tab 98.9 \tab 99 \tab 99.1 \tab 99.2 \tab 99.4 \tab 99.7 \tab 18.6 \tab 19 \tab 19.2 \tab 21.6 \tab 21.9 \tab 27 \tab 28.3 \tab 29.5 \tab 30.6 \tab 31.2 \tab 31.9 \tab 32.1 \tab 35.5 \tab 45.7 \tab 47.5 \tab 47.8 \tab 47.9 \tab 48.9 \tab 54.3 \tab 58.2 \tab 58.6 \tab 62.2 \tab 66.7 \tab 67.4 \tab 67.6 \tab 74.4 \tab 76.5 \tab 78.1 \tab 78.3 \tab 78.8 \tab 79.4 \tab 79.5 \tab 80 \tab 82.4 \tab 83.1 \tab 83.3 \tab 84.5 \tab 87.3 \tab 90.9 \tab 91.9 \tab 93.1 \tab 94.1 \tab 94.4 \tab 96 \tab 97 \tab 99.3 \tab don't know \tab 12.3 \tab 17.7 \tab 17.8 \tab 27.5 \tab 33.5 \tab 34.7 \tab 37.4 \tab 42.1 \tab 45.5 \tab 47.3 \tab 51.7 \tab 54.8 \tab 62.8 \tab 63.1 \tab 66.9 \tab 68.2 \tab 69.8 \tab 71.6 \tab 75.6 \tab 81.1 \tab 81.9 \tab 83.5 \tab 85.4 \tab 13.4 \tab 19.7 \tab 23.9 \tab 42.6 \tab 46.1 \tab 46.7 \tab 56.6 \tab 61.3 \tab 70.3 \tab 88.2 \tab 9.8 \tab 90.2 \tab 90.8 \tab 91.7 \tab 95 \tab 95.1 \tab 95.7 \tab 97.3 \tab 98.3 \tab 99.5 \tab 19.8 \tab 21.1 \tab 22.4 \tab 53.3 \tab 57.7 \tab 66.8 \tab 77.5 \tab 85.6 \tab no answer \tab 10.9 \tab 23.6 \tab 25.6 \tab 55.6 \tab 58.7 \tab 72.2 \tab 79.6 \tab 79.8 \tab 84.8 \tab 86.2 \tab 86.8 \tab 92.4 \tab 96.5 \tab 98.7 \tab 63.5 \tab 70.4 \tab 16.8 \tab 61 \tab 89.8 \tab 94.3 \tab 40.9 \tab 48.1 \tab 84.4 \tab 90.5 \tab 98.5 \tab 72.5 \tab 27.8 \tab 28 \tab 64.7 \tab 81 \tab 18 \tab 18.7 \tab 35.2 \tab 40.2 \tab 40.4 \tab 48.2 \tab 69.6 \tab 21.5 \tab 34.9 \tab 39.9 \tab 60.7 \tab 69.7 \tab 91.1 \tab 32.5 \tab 48.3 \tab 84.7 \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1351 \tab 7 \tab 2 \tab 15 \tab 1 \tab 1 \tab 19 \tab 2 \tab 3 \tab 72 \tab 4 \tab 12 \tab 14 \tab 113 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 64 \tab 5 \tab 2 \tab 11 \tab 39 \tab 31 \tab 1 \tab 3 \tab 6 \tab 2 \tab 1 \tab 5 \tab 6 \tab 15 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 5 \tab 11 \tab 1 \tab 2 \tab 1 \tab 39 \tab 1 \tab 1 \tab 7 \tab 2 \tab 21 \tab 27 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 73 \tab 1 \tab 2 \tab 12 \tab 26 \tab 6 \tab 1 \tab 1 \tab 51 \tab 1 \tab 1 \tab 11 \tab 28 \tab 4 \tab 1 \tab 3 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 8 \tab 1 \tab 8 \tab 1 \tab 2 \tab 2 \tab 7 \tab 7 \tab 14 \tab 2 \tab 1 \tab 37 \tab 2 \tab 3 \tab 3 \tab 75 \tab 68 \tab 7 \tab 1 \tab 138 \tab 2 \tab 15 \tab 6 \tab 3 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab 23 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 16 \tab 1 \tab 9 \tab 7 \tab 1 \tab 12 \tab 2 \tab 2 \tab 1 \tab 2 \tab 9 \tab 6 \tab 6 \tab 2 \tab 4 \tab 5 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 15 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 52 \tab 62 \tab 18 \tab 5 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1158 \tab 11 \tab 3 \tab 4 \tab - \tab 1 \tab 18 \tab 2 \tab 5 \tab 77 \tab 2 \tab 15 \tab 21 \tab 114 \tab 1 \tab 4 \tab 5 \tab - \tab - \tab - \tab - \tab 8 \tab 2 \tab 60 \tab 15 \tab 1 \tab 9 \tab 45 \tab 29 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 6 \tab 8 \tab 4 \tab 2 \tab - \tab 4 \tab 3 \tab 1 \tab 9 \tab 10 \tab 1 \tab 4 \tab - \tab 40 \tab - \tab - \tab 7 \tab 5 \tab 17 \tab 24 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 79 \tab - \tab 2 \tab 12 \tab 32 \tab 3 \tab - \tab - \tab 50 \tab - \tab - \tab 13 \tab 27 \tab 7 \tab - \tab 3 \tab 1 \tab 5 \tab 2 \tab - \tab 1 \tab 3 \tab 5 \tab - \tab 8 \tab 3 \tab 10 \tab 6 \tab - \tab 2 \tab 12 \tab 3 \tab 14 \tab 1 \tab - \tab 28 \tab 4 \tab 8 \tab 5 \tab 65 \tab 69 \tab 6 \tab 5 \tab 151 \tab 2 \tab 9 \tab 6 \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 18 \tab 2 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab - \tab 23 \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 9 \tab 2 \tab 14 \tab 6 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 6 \tab 7 \tab 5 \tab 6 \tab 5 \tab - \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 10 \tab - \tab 7 \tab 4 \tab 1 \tab - \tab 1 \tab 18 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 58 \tab 70 \tab 27 \tab 9 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1213 \tab 7 \tab 2 \tab 13 \tab - \tab - \tab 14 \tab - \tab 10 \tab 84 \tab 8 \tab 13 \tab 6 \tab 91 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 55 \tab 5 \tab 4 \tab 8 \tab 49 \tab 30 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab 8 \tab 3 \tab 9 \tab 5 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 7 \tab 9 \tab 2 \tab - \tab - \tab 42 \tab - \tab 2 \tab 8 \tab 2 \tab 19 \tab 20 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 74 \tab - \tab 2 \tab 8 \tab 17 \tab 1 \tab - \tab 1 \tab 42 \tab 1 \tab - \tab 10 \tab 31 \tab 13 \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab - \tab 14 \tab 3 \tab - \tab - \tab 11 \tab 2 \tab 18 \tab 3 \tab - \tab 33 \tab 3 \tab 5 \tab 3 \tab 60 \tab 60 \tab 5 \tab 2 \tab 125 \tab 3 \tab 14 \tab 10 \tab 7 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 12 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab 18 \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 10 \tab - \tab 10 \tab 5 \tab - \tab 12 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 10 \tab 6 \tab 6 \tab 7 \tab 1 \tab - \tab - \tab 5 \tab 4 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 15 \tab - \tab 1 \tab 7 \tab - \tab - \tab - \tab 13 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab 6 \tab 62 \tab 74 \tab 27 \tab 7 \tab 1 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1184 \tab 6 \tab 2 \tab 5 \tab 1 \tab 5 \tab 16 \tab 1 \tab 5 \tab 71 \tab 3 \tab 11 \tab 9 \tab 105 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 64 \tab 5 \tab - \tab 4 \tab 42 \tab 29 \tab 3 \tab 3 \tab 5 \tab - \tab - \tab 5 \tab 3 \tab 6 \tab 3 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 6 \tab 5 \tab - \tab 2 \tab - \tab 37 \tab - \tab - \tab 11 \tab 3 \tab 29 \tab 21 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 71 \tab - \tab 5 \tab 9 \tab 18 \tab 2 \tab - \tab - \tab 40 \tab 2 \tab - \tab 11 \tab 39 \tab 8 \tab 3 \tab 1 \tab - \tab 2 \tab 6 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 4 \tab 9 \tab - \tab - \tab 10 \tab 7 \tab 17 \tab - \tab - \tab 35 \tab 7 \tab 4 \tab 3 \tab 63 \tab 59 \tab 2 \tab 3 \tab 151 \tab 3 \tab 12 \tab 8 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 12 \tab 2 \tab 4 \tab 3 \tab 3 \tab 1 \tab 3 \tab 2 \tab 21 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 5 \tab 12 \tab 1 \tab 14 \tab 1 \tab - \tab 11 \tab 1 \tab - \tab 2 \tab 3 \tab 3 \tab 9 \tab 6 \tab 6 \tab 11 \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 11 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 10 \tab - \tab 4 \tab 1 \tab 4 \tab - \tab - \tab 6 \tab 57 \tab 63 \tab 19 \tab 10 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 6 \tab 3 \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1139 \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab 13 \tab - \tab 2 \tab 57 \tab 4 \tab 12 \tab 8 \tab 98 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 65 \tab 9 \tab 2 \tab 7 \tab 49 \tab 22 \tab 1 \tab - \tab 1 \tab - \tab - \tab 6 \tab 6 \tab 19 \tab 3 \tab 2 \tab 1 \tab 3 \tab 5 \tab - \tab 5 \tab 12 \tab 3 \tab 1 \tab 1 \tab 29 \tab 1 \tab - \tab 10 \tab 2 \tab 13 \tab 18 \tab 3 \tab - \tab 7 \tab 1 \tab - \tab 95 \tab - \tab 6 \tab 13 \tab 20 \tab 1 \tab 1 \tab - \tab 57 \tab - \tab - \tab 4 \tab 28 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 7 \tab 1 \tab 11 \tab 5 \tab 1 \tab - \tab 8 \tab 7 \tab 17 \tab 1 \tab - \tab 28 \tab 5 \tab 1 \tab 3 \tab 72 \tab 60 \tab 4 \tab 6 \tab 170 \tab 2 \tab 16 \tab 7 \tab 6 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 18 \tab 3 \tab 6 \tab 4 \tab 2 \tab 2 \tab - \tab 4 \tab 10 \tab 1 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 12 \tab 4 \tab 9 \tab 4 \tab - \tab 14 \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab 12 \tab 5 \tab 11 \tab - \tab - \tab - \tab 3 \tab 3 \tab 5 \tab - \tab 3 \tab - \tab 1 \tab 9 \tab 1 \tab 6 \tab 7 \tab 1 \tab 1 \tab - \tab 11 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 9 \tab 39 \tab 64 \tab 13 \tab 7 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1124 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab 23 \tab 1 \tab 2 \tab 45 \tab 4 \tab 7 \tab 11 \tab 93 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 67 \tab 7 \tab - \tab 11 \tab 40 \tab 23 \tab 4 \tab 1 \tab 4 \tab 1 \tab - \tab 7 \tab 9 \tab 7 \tab 6 \tab 2 \tab 1 \tab 3 \tab 5 \tab 2 \tab 4 \tab 13 \tab 1 \tab 1 \tab - \tab 28 \tab - \tab 4 \tab 9 \tab 2 \tab 13 \tab 14 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 72 \tab 1 \tab 5 \tab 1 \tab 15 \tab - \tab - \tab 1 \tab 57 \tab - \tab - \tab 6 \tab 39 \tab 3 \tab - \tab 5 \tab 1 \tab 3 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 7 \tab 3 \tab 11 \tab 3 \tab 1 \tab - \tab 19 \tab 3 \tab 9 \tab 1 \tab - \tab 30 \tab 7 \tab 7 \tab 5 \tab 78 \tab 58 \tab 5 \tab 1 \tab 135 \tab 2 \tab 13 \tab 14 \tab 6 \tab 4 \tab 1 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 11 \tab 6 \tab 4 \tab - \tab 2 \tab 2 \tab 7 \tab 1 \tab 28 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 8 \tab 19 \tab - \tab 21 \tab 1 \tab - \tab 11 \tab 1 \tab - \tab - \tab 3 \tab - \tab 11 \tab 8 \tab 6 \tab 12 \tab 5 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab - \tab 7 \tab - \tab 2 \tab 8 \tab 1 \tab - \tab - \tab 17 \tab - \tab 6 \tab 4 \tab 5 \tab - \tab - \tab 8 \tab 44 \tab 92 \tab 26 \tab 13 \tab - \tab 6 \tab 6 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2690 \tab 5 \tab 2 \tab 3 \tab - \tab 8 \tab 25 \tab - \tab 2 \tab 63 \tab 9 \tab 13 \tab 11 \tab 116 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 4 \tab 73 \tab 9 \tab 2 \tab 14 \tab 48 \tab 48 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab 4 \tab 8 \tab 7 \tab - \tab - \tab 2 \tab 2 \tab - \tab 6 \tab 9 \tab 1 \tab - \tab 1 \tab 35 \tab - \tab 2 \tab 13 \tab 3 \tab 13 \tab 20 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 92 \tab - \tab - \tab 5 \tab 24 \tab 1 \tab - \tab - \tab 48 \tab 1 \tab - \tab 9 \tab 40 \tab 5 \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab 7 \tab - \tab - \tab 2 \tab 11 \tab 7 \tab 13 \tab 1 \tab - \tab 34 \tab 3 \tab 6 \tab 3 \tab 72 \tab 58 \tab 7 \tab 1 \tab 145 \tab 1 \tab 17 \tab 9 \tab 6 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 15 \tab 4 \tab 2 \tab 2 \tab 2 \tab 1 \tab 7 \tab 4 \tab 12 \tab 3 \tab 4 \tab 2 \tab 3 \tab 8 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 13 \tab 6 \tab 8 \tab 6 \tab - \tab 19 \tab 1 \tab 1 \tab 1 \tab 3 \tab 8 \tab 7 \tab 11 \tab 9 \tab 11 \tab 9 \tab 1 \tab - \tab 6 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 12 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 24 \tab 3 \tab 6 \tab 2 \tab 5 \tab - \tab - \tab 7 \tab 53 \tab 85 \tab 23 \tab 17 \tab - \tab 4 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 4 \tab 1 \tab 2 \tab 4 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 730 \tab 2 \tab 3 \tab 4 \tab - \tab 1 \tab 14 \tab 1 \tab 7 \tab 40 \tab 4 \tab 11 \tab 9 \tab 84 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 51 \tab 9 \tab - \tab 4 \tab 17 \tab 21 \tab 2 \tab - \tab 2 \tab - \tab - \tab 7 \tab 5 \tab 7 \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 6 \tab 7 \tab - \tab 1 \tab - \tab 30 \tab - \tab - \tab 8 \tab - \tab 16 \tab 19 \tab - \tab - \tab 12 \tab - \tab 2 \tab 62 \tab - \tab 4 \tab 3 \tab 19 \tab 2 \tab - \tab - \tab 31 \tab - \tab - \tab 10 \tab 26 \tab 4 \tab 2 \tab 2 \tab - \tab 7 \tab 2 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 4 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 13 \tab 4 \tab 15 \tab 1 \tab - \tab 14 \tab 5 \tab 7 \tab - \tab 52 \tab 42 \tab 3 \tab 1 \tab 103 \tab 1 \tab 14 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab - \tab 14 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 9 \tab 5 \tab 7 \tab - \tab 2 \tab 2 \tab 2 \tab 5 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 11 \tab 3 \tab 8 \tab 6 \tab - \tab 12 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 8 \tab - \tab 3 \tab 6 \tab 1 \tab - \tab 1 \tab 14 \tab 1 \tab 4 \tab 3 \tab 4 \tab 2 \tab - \tab 4 \tab 39 \tab 49 \tab 18 \tab 4 \tab 1 \tab 4 \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 677 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 11 \tab 3 \tab 2 \tab 50 \tab 7 \tab 5 \tab 4 \tab 66 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 46 \tab 13 \tab - \tab 4 \tab 33 \tab 28 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 7 \tab 5 \tab 10 \tab 3 \tab - \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 8 \tab - \tab 1 \tab - \tab 24 \tab - \tab 1 \tab 3 \tab 4 \tab 11 \tab 19 \tab - \tab 1 \tab 5 \tab - \tab - \tab 64 \tab - \tab 6 \tab 6 \tab 16 \tab 2 \tab 1 \tab - \tab 52 \tab - \tab - \tab 12 \tab 27 \tab 5 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 6 \tab - \tab - \tab 7 \tab 1 \tab 9 \tab 7 \tab 1 \tab 2 \tab 9 \tab 3 \tab 18 \tab - \tab - \tab 28 \tab 3 \tab 5 \tab 4 \tab 45 \tab 42 \tab 2 \tab - \tab 102 \tab 2 \tab 6 \tab 12 \tab 7 \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 13 \tab 4 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 15 \tab 1 \tab 3 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 10 \tab - \tab 13 \tab 2 \tab 1 \tab 10 \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 4 \tab 8 \tab 6 \tab 6 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 11 \tab - \tab 7 \tab 4 \tab 1 \tab 1 \tab - \tab 22 \tab - \tab 8 \tab - \tab 5 \tab - \tab - \tab 10 \tab 48 \tab 59 \tab 28 \tab 7 \tab 1 \tab 6 \tab 10 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 647 \tab 6 \tab 4 \tab 4 \tab - \tab 2 \tab 11 \tab - \tab 5 \tab 42 \tab 6 \tab 13 \tab 8 \tab 74 \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 59 \tab 8 \tab 1 \tab 13 \tab 44 \tab 12 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 3 \tab 6 \tab 4 \tab 2 \tab - \tab 1 \tab 5 \tab 1 \tab 3 \tab 6 \tab - \tab - \tab 1 \tab 21 \tab 1 \tab - \tab 5 \tab 1 \tab 14 \tab 23 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 36 \tab - \tab 1 \tab 5 \tab 20 \tab 1 \tab - \tab - \tab 36 \tab 1 \tab - \tab 3 \tab 25 \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab 13 \tab 2 \tab 8 \tab 3 \tab - \tab - \tab 17 \tab 7 \tab 6 \tab 1 \tab - \tab 31 \tab 2 \tab 1 \tab 2 \tab 53 \tab 49 \tab 2 \tab 2 \tab 85 \tab 2 \tab 10 \tab 11 \tab 4 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 12 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab 11 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 6 \tab - \tab - \tab 3 \tab 5 \tab 16 \tab - \tab 7 \tab 2 \tab - \tab 9 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 12 \tab 7 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab 4 \tab 7 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 5 \tab - \tab 5 \tab 7 \tab 4 \tab 2 \tab - \tab 17 \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab 45 \tab 46 \tab 21 \tab 7 \tab - \tab 5 \tab 6 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 5 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 20 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 866 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 17 \tab 1 \tab 1 \tab 45 \tab 5 \tab 9 \tab 11 \tab 80 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 67 \tab 11 \tab 3 \tab 7 \tab 61 \tab 19 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 9 \tab 1 \tab 5 \tab 4 \tab 1 \tab - \tab 2 \tab 7 \tab 3 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab 41 \tab 1 \tab - \tab 11 \tab 2 \tab 9 \tab 24 \tab 1 \tab 2 \tab 12 \tab 1 \tab 2 \tab 73 \tab - \tab 5 \tab 5 \tab 17 \tab 2 \tab 1 \tab 1 \tab 42 \tab 1 \tab - \tab 8 \tab 40 \tab 3 \tab 2 \tab 4 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 5 \tab 1 \tab 17 \tab 3 \tab - \tab 4 \tab 16 \tab 10 \tab 7 \tab 1 \tab 1 \tab 40 \tab 6 \tab 7 \tab 2 \tab 51 \tab 54 \tab 4 \tab 2 \tab 144 \tab 3 \tab 15 \tab 6 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 10 \tab 1 \tab 5 \tab 3 \tab 5 \tab 5 \tab 9 \tab 5 \tab 22 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab 8 \tab 10 \tab 2 \tab 18 \tab 3 \tab - \tab 18 \tab - \tab - \tab 1 \tab 4 \tab 4 \tab 5 \tab 7 \tab 6 \tab 11 \tab 4 \tab - \tab - \tab 4 \tab 10 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 12 \tab 1 \tab 5 \tab 3 \tab 3 \tab - \tab - \tab 22 \tab 1 \tab 7 \tab 2 \tab 5 \tab - \tab - \tab 8 \tab 60 \tab 51 \tab 32 \tab 16 \tab - \tab 3 \tab 8 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 4 \tab 3 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 6 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 875 \tab 6 \tab 4 \tab 3 \tab - \tab 1 \tab 21 \tab - \tab 4 \tab 43 \tab 7 \tab 10 \tab 10 \tab 128 \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab 3 \tab 98 \tab 6 \tab - \tab 5 \tab 46 \tab 22 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab 9 \tab 4 \tab 13 \tab 10 \tab 2 \tab - \tab 3 \tab 5 \tab - \tab 4 \tab 5 \tab 1 \tab 4 \tab 1 \tab 43 \tab - \tab - \tab 15 \tab 7 \tab 20 \tab 22 \tab - \tab 1 \tab 18 \tab 1 \tab 2 \tab 78 \tab - \tab 3 \tab 6 \tab 17 \tab 3 \tab 2 \tab 2 \tab 52 \tab 1 \tab - \tab 10 \tab 45 \tab 2 \tab 2 \tab 2 \tab - \tab 5 \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab 4 \tab 3 \tab 16 \tab 3 \tab - \tab - \tab 17 \tab 8 \tab 9 \tab - \tab 1 \tab 32 \tab 9 \tab 3 \tab 10 \tab 55 \tab 69 \tab 4 \tab 2 \tab 125 \tab 3 \tab 21 \tab 19 \tab 4 \tab - \tab 1 \tab 5 \tab 2 \tab 6 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 25 \tab 7 \tab 9 \tab 3 \tab 1 \tab 2 \tab 9 \tab 4 \tab 14 \tab 1 \tab 2 \tab 2 \tab 1 \tab 8 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 8 \tab 20 \tab 2 \tab 10 \tab 7 \tab 1 \tab 21 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 15 \tab 16 \tab 12 \tab 20 \tab 5 \tab - \tab - \tab 1 \tab 6 \tab 4 \tab 4 \tab - \tab 1 \tab - \tab 16 \tab - \tab 8 \tab 5 \tab 2 \tab - \tab 1 \tab 20 \tab 4 \tab 10 \tab 1 \tab 4 \tab 1 \tab 1 \tab 10 \tab 103 \tab 71 \tab 32 \tab 17 \tab 4 \tab 1 \tab 10 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 6 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 691 \tab 4 \tab 4 \tab - \tab 1 \tab 1 \tab 14 \tab - \tab 4 \tab 29 \tab 5 \tab 16 \tab 6 \tab 71 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 72 \tab 11 \tab - \tab 3 \tab 43 \tab 6 \tab 5 \tab 1 \tab 4 \tab 1 \tab - \tab 11 \tab 4 \tab 7 \tab 3 \tab 4 \tab - \tab - \tab 12 \tab 3 \tab 3 \tab 8 \tab - \tab - \tab 1 \tab 50 \tab - \tab 1 \tab 17 \tab 3 \tab 19 \tab 23 \tab - \tab - \tab 4 \tab - \tab 2 \tab 75 \tab - \tab 1 \tab 10 \tab 18 \tab 1 \tab - \tab 2 \tab 37 \tab - \tab - \tab 4 \tab 48 \tab 3 \tab 1 \tab 2 \tab - \tab 5 \tab 2 \tab 3 \tab - \tab 3 \tab 6 \tab 1 \tab 4 \tab - \tab 17 \tab 5 \tab - \tab 2 \tab 16 \tab 2 \tab 3 \tab 2 \tab - \tab 26 \tab 3 \tab 4 \tab 9 \tab 63 \tab 40 \tab 1 \tab 2 \tab 127 \tab 3 \tab 12 \tab 5 \tab 10 \tab 2 \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 16 \tab 7 \tab 8 \tab - \tab 5 \tab 1 \tab 6 \tab 2 \tab 12 \tab 2 \tab - \tab 2 \tab 3 \tab 6 \tab 5 \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 7 \tab 16 \tab - \tab 6 \tab 6 \tab 1 \tab 11 \tab - \tab 1 \tab 2 \tab 6 \tab 6 \tab 9 \tab 17 \tab 17 \tab 22 \tab 3 \tab - \tab 1 \tab 6 \tab 8 \tab 4 \tab 3 \tab - \tab 2 \tab 1 \tab 12 \tab 1 \tab 5 \tab 9 \tab 1 \tab 1 \tab - \tab 32 \tab 1 \tab 6 \tab - \tab 5 \tab 2 \tab - \tab 3 \tab 65 \tab 47 \tab 32 \tab 13 \tab 1 \tab 4 \tab 7 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 7 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1266 \tab 7 \tab 8 \tab 1 \tab - \tab 4 \tab 19 \tab - \tab - \tab 71 \tab 7 \tab 8 \tab 14 \tab 103 \tab 1 \tab 1 \tab 9 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 88 \tab 15 \tab 1 \tab 12 \tab 76 \tab 28 \tab 6 \tab 3 \tab 5 \tab 1 \tab - \tab 7 \tab 21 \tab 15 \tab 6 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 10 \tab - \tab - \tab - \tab 55 \tab - \tab - \tab 19 \tab 3 \tab 11 \tab 49 \tab - \tab 2 \tab 20 \tab 2 \tab 1 \tab 89 \tab - \tab 2 \tab 8 \tab 20 \tab 9 \tab - \tab 1 \tab 77 \tab 2 \tab - \tab 4 \tab 53 \tab 4 \tab 2 \tab 5 \tab - \tab 4 \tab 3 \tab - \tab - \tab 6 \tab 5 \tab 2 \tab 13 \tab - \tab 27 \tab 6 \tab - \tab - \tab 26 \tab 11 \tab 17 \tab - \tab - \tab 49 \tab 10 \tab 10 \tab 13 \tab 114 \tab 158 \tab 9 \tab 2 \tab 235 \tab 3 \tab 29 \tab 22 \tab 4 \tab 2 \tab 2 \tab 3 \tab 2 \tab 9 \tab - \tab - \tab 3 \tab 1 \tab 4 \tab 1 \tab 22 \tab 2 \tab 4 \tab - \tab 1 \tab 3 \tab 7 \tab 3 \tab 40 \tab 2 \tab 4 \tab 4 \tab 6 \tab 12 \tab 7 \tab 3 \tab 2 \tab 6 \tab 1 \tab 3 \tab 3 \tab 11 \tab 19 \tab 4 \tab 19 \tab 9 \tab - \tab 21 \tab 2 \tab - \tab 3 \tab 2 \tab 3 \tab 9 \tab 20 \tab 16 \tab 17 \tab 8 \tab - \tab - \tab 10 \tab 5 \tab 4 \tab 2 \tab - \tab - \tab 3 \tab 30 \tab 2 \tab 12 \tab 15 \tab 3 \tab - \tab 1 \tab 27 \tab 4 \tab 15 \tab 5 \tab 12 \tab - \tab 1 \tab 14 \tab 115 \tab 146 \tab 32 \tab 17 \tab 3 \tab 18 \tab 12 \tab 1 \tab - \tab 4 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 3 \tab 3 \tab 4 \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 4 \tab 8 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 4 \tab 2 \tab 6 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1040 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 21 \tab - \tab 1 \tab 47 \tab 8 \tab 9 \tab 14 \tab 118 \tab 3 \tab - \tab 5 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 96 \tab 9 \tab 1 \tab 19 \tab 60 \tab 39 \tab 8 \tab - \tab 1 \tab 1 \tab - \tab 8 \tab 14 \tab 6 \tab 5 \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 15 \tab 2 \tab 1 \tab - \tab 51 \tab - \tab - \tab 12 \tab 2 \tab 20 \tab 31 \tab 1 \tab - \tab 30 \tab - \tab 1 \tab 86 \tab - \tab 4 \tab 9 \tab 23 \tab 5 \tab - \tab 1 \tab 59 \tab 1 \tab 1 \tab 15 \tab 46 \tab 6 \tab 1 \tab 6 \tab - \tab 9 \tab 5 \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab 15 \tab 2 \tab 33 \tab 7 \tab 2 \tab 1 \tab 19 \tab 11 \tab 12 \tab - \tab 2 \tab 36 \tab 8 \tab 5 \tab 8 \tab 95 \tab 120 \tab 7 \tab 4 \tab 177 \tab 2 \tab 20 \tab 40 \tab 7 \tab 1 \tab 3 \tab 4 \tab 2 \tab 8 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 15 \tab 9 \tab 8 \tab 4 \tab 5 \tab 1 \tab 5 \tab 3 \tab 37 \tab 2 \tab 5 \tab 2 \tab 1 \tab 10 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 9 \tab 29 \tab 1 \tab 18 \tab 2 \tab 2 \tab 17 \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 13 \tab 9 \tab 19 \tab 16 \tab 6 \tab - \tab - \tab 9 \tab 7 \tab 4 \tab 4 \tab 3 \tab - \tab 3 \tab 27 \tab - \tab 6 \tab 12 \tab 2 \tab - \tab 1 \tab 38 \tab 1 \tab 11 \tab 2 \tab 12 \tab - \tab - \tab 6 \tab 110 \tab 115 \tab 42 \tab 15 \tab 4 \tab 8 \tab 12 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab 4 \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 4 \tab 9 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 8 \tab 2 \tab 4 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 5 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 46039 \tab 74 \tab 41 \tab 61 \tab 3 \tab 32 \tab 256 \tab 11 \tab 53 \tab 836 \tab 83 \tab 164 \tab 156 \tab 1454 \tab 18 \tab 20 \tab 54 \tab 3 \tab 11 \tab 15 \tab 11 \tab 47 \tab 28 \tab 1025 \tab 137 \tab 17 \tab 131 \tab 692 \tab 387 \tab 47 \tab 24 \tab 45 \tab 8 \tab 1 \tab 100 \tab 94 \tab 141 \tab 67 \tab 24 \tab 7 \tab 29 \tab 65 \tab 20 \tab 67 \tab 134 \tab 13 \tab 18 \tab 7 \tab 565 \tab 4 \tab 11 \tab 155 \tab 41 \tab 245 \tab 354 \tab 9 \tab 15 \tab 138 \tab 14 \tab 18 \tab 1119 \tab 2 \tab 48 \tab 112 \tab 302 \tab 39 \tab 6 \tab 10 \tab 731 \tab 11 \tab 2 \tab 130 \tab 542 \tab 73 \tab 21 \tab 41 \tab 4 \tab 63 \tab 44 \tab 12 \tab 5 \tab 41 \tab 39 \tab 20 \tab 106 \tab 19 \tab 198 \tab 62 \tab 8 \tab 16 \tab 211 \tab 92 \tab 189 \tab 14 \tab 5 \tab 481 \tab 77 \tab 76 \tab 73 \tab 1013 \tab 1006 \tab 68 \tab 34 \tab 2113 \tab 34 \tab 223 \tab 183 \tab 76 \tab 21 \tab 23 \tab 35 \tab 28 \tab 49 \tab 3 \tab 1 \tab 21 \tab 13 \tab 29 \tab 5 \tab 225 \tab 55 \tab 68 \tab 27 \tab 35 \tab 30 \tab 79 \tab 40 \tab 293 \tab 16 \tab 24 \tab 29 \tab 30 \tab 81 \tab 34 \tab 26 \tab 23 \tab 25 \tab 12 \tab 23 \tab 24 \tab 83 \tab 222 \tab 26 \tab 184 \tab 67 \tab 6 \tab 205 \tab 10 \tab 8 \tab 13 \tab 44 \tab 46 \tab 124 \tab 143 \tab 121 \tab 160 \tab 57 \tab 2 \tab 3 \tab 66 \tab 67 \tab 38 \tab 32 \tab 9 \tab 7 \tab 13 \tab 188 \tab 6 \tab 78 \tab 95 \tab 24 \tab 6 \tab 6 \tab 300 \tab 21 \tab 102 \tab 25 \tab 67 \tab 7 \tab 4 \tab 104 \tab 950 \tab 1094 \tab 390 \tab 164 \tab 21 \tab 69 \tab 102 \tab 4 \tab 7 \tab 13 \tab 14 \tab 12 \tab 12 \tab 6 \tab 11 \tab 1 \tab 9 \tab 43 \tab 6 \tab 8 \tab 8 \tab 4 \tab 7 \tab 1 \tab 19 \tab 20 \tab 3 \tab 16 \tab 3 \tab 1 \tab 23 \tab 13 \tab 29 \tab 28 \tab 8 \tab 16 \tab 12 \tab 24 \tab 10 \tab 17 \tab 59 \tab 15 \tab 6 \tab 15 \tab 28 \tab 4 \tab 51 \tab 7 \tab 20 \tab 17 \tab 10 \tab 4 \tab 18 \tab 18 \tab 1 \tab 4 \tab 1 \tab 4 \tab 7 \tab 15 \tab 6 \tab 1 \tab 2 \tab 11 \tab 10 \tab 19 \tab 7 \tab 11 \tab 55 \tab 4 \tab 5 \tab 27 \tab 10 \tab 6 \tab 6 \tab 9 \tab 5 \tab 4 \tab 7 \tab 13 \tab 3 \tab 12 \tab 3 \tab 7 \tab 10 \tab 9 \tab 1 \tab 3 \tab 5 \tab 9 \tab 10 \tab 7 \tab 7 \tab 7 \tab 12 \tab 23 \tab 11 \tab 1 \tab 3 \tab 1 \tab 2 \tab 11 \tab 2 \tab 7 \tab 6 \tab 3 \tab 1 \tab 2 \tab 1 \tab 6 \tab 13 \tab 3 \tab 1 \tab 2 \tab 8 \tab 2 \tab 24 \tab 7 \tab 1 \tab 1 \tab 5 \tab 3 \tab 2 \tab 2 \tab 9 \tab 3 \tab 4 \tab 4 \tab 1 \tab 6 \tab 2 \tab 3 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{masei10}},
\code{\link{masei10inc}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}}
}
\concept{Occupation}
\keyword{variable}
