% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{matesex}
\alias{matesex}
\title{Was 1 of r's partner's spouse or regular}
\source{
General Social Survey https://gss.norc.org
}
\description{
matesex
}
\details{
Question Was one of the partners your husband or wife or regular sexual partner?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5050/vshow}{GSS Data Explorer page for \code{matesex}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab not available in this year \tab Total \cr
   1988 \tab 409 \tab 973 \tab 88 \tab 11 \tab - \tab - \tab 1481 \cr
   1989 \tab 444 \tab 988 \tab 90 \tab 15 \tab - \tab - \tab 1537 \cr
   1990 \tab 429 \tab 845 \tab 75 \tab 23 \tab - \tab - \tab 1372 \cr
   1991 \tab 511 \tab 884 \tab 84 \tab 38 \tab - \tab - \tab 1517 \cr
   1993 \tab 427 \tab 1055 \tab 80 \tab 44 \tab - \tab - \tab 1606 \cr
   1994 \tab 809 \tab 1918 \tab 185 \tab 80 \tab - \tab - \tab 2992 \cr
   1996 \tab 756 \tab 1902 \tab 182 \tab 64 \tab - \tab - \tab 2904 \cr
   1998 \tab 895 \tab 1699 \tab 186 \tab 52 \tab - \tab - \tab 2832 \cr
   2000 \tab 938 \tab 1615 \tab 174 \tab 90 \tab - \tab - \tab 2817 \cr
   2002 \tab 1000 \tab 1540 \tab 181 \tab 43 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1013 \tab 1637 \tab 128 \tab 34 \tab - \tab - \tab 2812 \cr
   2006 \tab 2693 \tab 1609 \tab 194 \tab 14 \tab - \tab - \tab 4510 \cr
   2008 \tab 659 \tab 1200 \tab 143 \tab 20 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 632 \tab 1274 \tab 117 \tab 20 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 643 \tab 1193 \tab 106 \tab 32 \tab - \tab - \tab 1974 \cr
   2014 \tab 727 \tab 1593 \tab 167 \tab 51 \tab - \tab - \tab 2538 \cr
   2016 \tab 1506 \tab 1194 \tab 147 \tab 20 \tab - \tab - \tab 2867 \cr
   2018 \tab 1301 \tab 931 \tab 101 \tab 15 \tab - \tab - \tab 2348 \cr
   2021 \tab 2307 \tab 1567 \tab 130 \tab 20 \tab 8 \tab - \tab 4032 \cr
   2022 \tab 2409 \tab 1000 \tab 109 \tab 17 \tab 9 \tab - \tab 3544 \cr
   Total \tab 20508 \tab 26617 \tab 2667 \tab 703 \tab 20 \tab 0 \tab 50515 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{frndsex}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{idu30}},
\code{\link{nocondom}},
\code{\link{nodocidu}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexornt}},
\code{\link{sexornt1}},
\code{\link{sexpaid}},
\code{\link{sexsex}},
\code{\link{sexsex5}}

Other Sex: 
\code{\link{acpttg}},
\code{\link{acqntsex}},
\code{\link{attractd}},
\code{\link{chngdocs}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{condom}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{family16sex}},
\code{\link{frndsex}},
\code{\link{harsexcl}},
\code{\link{harsexjb}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kidxfilm}},
\code{\link{knwclsex}},
\code{\link{numclerg}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{othersex}},
\code{\link{ownclerg}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{premars1}},
\code{\link{premarsx}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{relgendr}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sellsex}},
\code{\link{sex}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sexeduc}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexnow2}},
\code{\link{sexornt}},
\code{\link{sexsex}},
\code{\link{sexsex18}},
\code{\link{sexsex5}},
\code{\link{sexunion}},
\code{\link{spvtrfair}},
\code{\link{teensex}},
\code{\link{teensex1}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{xmarsex}},
\code{\link{xmarsex1}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{High Risk Behaviors}
\concept{Sex}
\keyword{variable}
