% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{mavisit}
\alias{mavisit}
\title{How often visit mother}
\source{
General Social Survey https://gss.norc.org
}
\description{
mavisit
}
\details{
Question Is your mother still alive?
A. How often do you see or visit your mother?
}
\section{Values}{

\itemize{
\item \code{1} lives in hshld
\item \code{2} daily
\item \code{3} several x week
\item \code{4} once a week
\item \code{5} once a month
\item \code{6} several x year
\item \code{7} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
   Full \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4128/vshow}{GSS Data Explorer page for \code{mavisit}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab lives in hshld \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab don't know \tab no answer \tab Total \cr
   1986 \tab 572 \tab 105 \tab 71 \tab 139 \tab 150 \tab 138 \tab 164 \tab 125 \tab 1 \tab 5 \tab 1470 \cr
   1994 \tab 2070 \tab 77 \tab 56 \tab 107 \tab 159 \tab 145 \tab 160 \tab 160 \tab - \tab 58 \tab 2992 \cr
   Total \tab 2642 \tab 182 \tab 127 \tab 246 \tab 309 \tab 283 \tab 324 \tab 285 \tab 1 \tab 63 \tab 4462 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{macall}},
\code{\link{malive}},
\code{\link{matime}},
\code{\link{palive}},
\code{\link{peocntct}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
