% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_457.R
\name{mavisit1}
\alias{mavisit1}
\title{How often does r visit mother}
\source{
General Social Survey https://gss.norc.org
}
\description{
mavisit1
}
\details{
Question And what about your mother? How often do you see or visit her?
}
\section{Values}{

\itemize{
\item \verb{[1]} she lives in the same household as i do
\item \verb{[2]} daily
\item \verb{[3]} at least several times a week
\item \verb{[4]} at least once a week
\item \verb{[5]} at least once a month
\item \verb{[6]} several times a year
\item \verb{[7]} less often
\item \verb{[8]} never
\item \verb{[9]} my mother is no longer alive
\item \verb{[10]} i dont know where mother lives
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4194/vshow}{GSS Data Explorer page for \code{mavisit1}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab she lives in the same household as i do \tab daily \tab at least several times a week \tab at least once a week \tab at least once a month \tab several times a year \tab less often \tab never \tab my mother is no longer alive \tab i dont know where mother lives \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 66 \tab 71 \tab 128 \tab 139 \tab 111 \tab 125 \tab 85 \tab 19 \tab 400 \tab 1 \tab 4 \tab 2765 \cr
   Total \tab 1616 \tab 66 \tab 71 \tab 128 \tab 139 \tab 111 \tab 125 \tab 85 \tab 19 \tab 400 \tab 1 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{cousins}},
\code{\link{kidcall}},
\code{\link{macall1}},
\code{\link{niecenep}},
\code{\link{pacall1}},
\code{\link{parslaw}},
\code{\link{pavisit1}},
\code{\link{sibinlaw}},
\code{\link{uncaunts}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
