% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{mawk16}
\alias{mawk16}
\title{Mothers employment when r growing up}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawk16
}
\details{
Question Did your mother ever work for pay for as long as a year, after she was married?
C. Did she work for as long as a year around the time you were 16?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/91/vshow}{GSS Data Explorer page for \code{mawk16}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab 913 \tab 399 \tab 185 \tab 6 \tab 1 \tab 1504 \cr
   1974 \tab 882 \tab 412 \tab 186 \tab 2 \tab 2 \tab 1484 \cr
   1975 \tab 794 \tab 447 \tab 226 \tab 4 \tab 19 \tab 1490 \cr
   1976 \tab 802 \tab 438 \tab 233 \tab 5 \tab 21 \tab 1499 \cr
   1977 \tab 877 \tab 419 \tab 201 \tab 2 \tab 31 \tab 1530 \cr
   1978 \tab 776 \tab 491 \tab 253 \tab 2 \tab 10 \tab 1532 \cr
   1980 \tab 743 \tab 447 \tab 259 \tab 2 \tab 17 \tab 1468 \cr
   1982 \tab 944 \tab 631 \tab 216 \tab 2 \tab 67 \tab 1860 \cr
   1983 \tab 731 \tab 584 \tab 273 \tab 4 \tab 7 \tab 1599 \cr
   Total \tab 7462 \tab 4268 \tab 2032 \tab 29 \tab 175 \tab 13966 \cr
}
}

\seealso{
Other Core: 
\code{\link{born}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{incom16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{parborn}}
}
\concept{Core}
\keyword{variable}
