% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{mawkborn}
\alias{mawkborn}
\title{Mothers employment after r was born}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawkborn
}
\details{
Question Did your mother ever work for pay for as long as a year, after she was married?
B. Did she work as long as a year after you were born and before you started first grade?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/90/vshow}{GSS Data Explorer page for \code{mawkborn}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1987 \tab 793 \tab 476 \tab 501 \tab 35 \tab 14 \tab 1819 \cr
   1988 \tab 637 \tab 374 \tab 430 \tab 36 \tab 4 \tab 1481 \cr
   1989 \tab 689 \tab 339 \tab 478 \tab 26 \tab 5 \tab 1537 \cr
   1990 \tab 571 \tab 370 \tab 396 \tab 27 \tab 8 \tab 1372 \cr
   1991 \tab 644 \tab 369 \tab 464 \tab 31 \tab 9 \tab 1517 \cr
   1993 \tab 652 \tab 418 \tab 486 \tab 32 \tab 18 \tab 1606 \cr
   1994 \tab 2586 \tab 182 \tab 200 \tab 17 \tab 7 \tab 2992 \cr
   Total \tab 6572 \tab 2528 \tab 2955 \tab 204 \tab 65 \tab 12324 \cr
}
}

\seealso{
Other Core: 
\code{\link{born}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{incom16}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{parborn}}
}
\concept{Core}
\keyword{variable}
