% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{mawrkslf}
\alias{mawrkslf}
\title{Mother (or youngest same-sex parent) self-emp. or worked for somebody}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawrkslf
}
\details{
Question A. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up?  That is, what was her job called?
OCCUPATION: (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What did she actually do in that job? Tell me, what were some of her main duties?
C. What kind of place did she work for?
INDUSTRY: (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was she self-employed, or did she work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/47/vshow}{GSS Data Explorer page for \code{mawrkslf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab self-employed \tab someone else \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1297 \tab 211 \tab 1445 \tab 1 \tab 38 \tab 2992 \cr
   1996 \tab 1124 \tab 205 \tab 1531 \tab 1 \tab 43 \tab 2904 \cr
   1998 \tab 1182 \tab 174 \tab 1447 \tab 2 \tab 27 \tab 2832 \cr
   2000 \tab 1150 \tab 214 \tab 1424 \tab 6 \tab 23 \tab 2817 \cr
   2002 \tab 1103 \tab 181 \tab 1465 \tab - \tab 16 \tab 2765 \cr
   2004 \tab 1076 \tab 203 \tab 1526 \tab 5 \tab 2 \tab 2812 \cr
   2006 \tab 2637 \tab 206 \tab 1635 \tab 2 \tab 30 \tab 4510 \cr
   2008 \tab 710 \tab 140 \tab 1166 \tab 5 \tab 2 \tab 2023 \cr
   2010 \tab 663 \tab 166 \tab 1212 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 602 \tab 159 \tab 1210 \tab 2 \tab 1 \tab 1974 \cr
   2014 \tab 848 \tab 190 \tab 1496 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 819 \tab 225 \tab 1813 \tab 5 \tab 5 \tab 2867 \cr
   2018 \tab 672 \tab 171 \tab 1500 \tab 3 \tab 2 \tab 2348 \cr
   2021 \tab 1233 \tab 287 \tab 2510 \tab - \tab 2 \tab 4032 \cr
   2022 \tab 966 \tab 261 \tab 2316 \tab - \tab 1 \tab 3544 \cr
   Total \tab 16082 \tab 2993 \tab 23696 \tab 39 \tab 192 \tab 43002 \cr
}
}

\seealso{
Other Core: 
\code{\link{maind80}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{paind10}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{papres10}},
\code{\link{papres105plus}}
}
\concept{Core}
\keyword{variable}
