% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_612.R
\name{mawrkslf2}
\alias{mawrkslf2}
\title{At this job, was mother (or youngest same-sex parent) an employee, self-employed without employees, or self-employed with employees?}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawrkslf2
}
\details{
Question At this job, was your mother an employee, self-employed without employees, or self-employed with employees?
}
\section{Values}{

\itemize{
\item \verb{[1]} employee
\item \verb{[2]} self-employed without employees
\item \verb{[3]} self-employed with employees
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7430/vshow}{GSS Data Explorer page for \code{mawrkslf2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab employee \tab self-employed without employees \tab self-employed with employees \tab no answer \tab Total \cr
   2021 \tab 1233 \tab 2510 \tab 164 \tab 123 \tab 2 \tab 4032 \cr
   2022 \tab 966 \tab 2316 \tab 171 \tab 90 \tab 1 \tab 3544 \cr
   Total \tab 2199 \tab 4826 \tab 335 \tab 213 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
