% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_226.R
\name{mcsds3}
\alias{mcsds3}
\title{At times i have really insisted on having things my own way}
\source{
General Social Survey https://gss.norc.org
}
\description{
mcsds3
}
\details{
Question Next are a number of statements concerning personal attitudes and traits. Please tell me whether the statement is true or false as it pertains to you personally: At times I have really insisted on having things my own way.
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7151/vshow}{GSS Data Explorer page for \code{mcsds3}}.
}

\keyword{variable}
