% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_226.R
\name{mcsds7}
\alias{mcsds7}
\title{There have been occasions when i took advantage of someone}
\source{
General Social Survey https://gss.norc.org
}
\description{
mcsds7
}
\details{
Question Next are a number of statements concerning personal attitudes and traits. Please tell me whether the statement is true or false as it pertains to you personally: There have been occasions when I took advantage of someone.
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7155/vshow}{GSS Data Explorer page for \code{mcsds7}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 363 \tab 777 \tab 6 \tab 27 \tab 2348 \cr
   Total \tab 1175 \tab 363 \tab 777 \tab 6 \tab 27 \tab 2348 \cr
}
}

\keyword{variable}
