% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{mebear}
\alias{mebear}
\title{Can husband refuse to sire children}
\source{
General Social Survey https://gss.norc.org
}
\description{
mebear
}
\details{
Question If the wife in a family wants children, but the husband decides that he does not want any children, is it all right for the husband to refuse to have children.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1696/vshow}{GSS Data Explorer page for \code{mebear}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2199 \tab 366 \tab 232 \tab 106 \tab 1 \tab 2904 \cr
   Total \tab 2199 \tab 366 \tab 232 \tab 106 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{fehire}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{hmemaker}},
\code{\link{wrkclass}}
}
\concept{Gender}
\keyword{variable}
