% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_333.R
\name{medaddct}
\alias{medaddct}
\title{Med for mental health problem are addictive}
\source{
General Social Survey https://gss.norc.org
}
\description{
medaddct
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors that are intended to help people who are having problems with their emotions, nerves or their mental health.
B. These medications are addictive.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree somewhat
\item \verb{[3]} disagree somewhat
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2975/vshow}{GSS Data Explorer page for \code{medaddct}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree somewhat \tab disagree somewhat \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 472 \tab 564 \tab 181 \tab 53 \tab 108 \tab 15 \tab 2765 \cr
   Total \tab 1372 \tab 472 \tab 564 \tab 181 \tab 53 \tab 108 \tab 15 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{solveown}}
}
\concept{Mental Health II}
\keyword{variable}
