% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_373.R
\name{medagrsc}
\alias{medagrsc}
\title{Agreement by med researchers on importance of stem cell rsch}
\source{
General Social Survey https://gss.norc.org
}
\description{
medagrsc
}
\details{
Question Please look at Card B7. (On a scale of 1 to 5, where 1 means "Near complete agreement" and 5 means "No agreement at all") to what extent do medical researchers agree on the importance of stem cells for research?
}
\section{Values}{

\itemize{
\item \verb{[1]} near complete agreement
\item \verb{[5]} no agreement at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3369/vshow}{GSS Data Explorer page for \code{medagrsc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab near complete agreement \tab 2 \tab 3 \tab 4 \tab no agreement at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 189 \tab 316 \tab 260 \tab 42 \tab 44 \tab 74 \tab 3 \tab 4510 \cr
   2010 \tab 1569 \tab 119 \tab 157 \tab 110 \tab 26 \tab 18 \tab 40 \tab 5 \tab 2044 \cr
   Total \tab 5151 \tab 308 \tab 473 \tab 370 \tab 68 \tab 62 \tab 114 \tab 8 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{medbstsc}},
\code{\link{medinfsc}},
\code{\link{polbstsc}},
\code{\link{polinfsc}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{scresrch}},
\code{\link{txbiz}},
\code{\link{txeco}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
