% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{medbest}
\alias{medbest}
\title{How likely to get best treatment avail in us}
\source{
General Social Survey https://gss.norc.org
}
\description{
medbest
}
\details{
Question How likely is it that if you become seriously ill, you would get or not get the best treatment available in the United States?
}
\section{Values}{

\itemize{
\item \code{1} it's certain r would get
\item \code{2} it's likely r would get
\item \code{3} equal chance of getting or not getting
\item \code{4} it's likely r would not get
\item \code{5} it's certain r would not get
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7830/vshow}{GSS Data Explorer page for \code{medbest}}.
}

\keyword{variable}
