% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_375.R
\name{medbstgm}
\alias{medbstgm}
\title{Best policy on gmf vs narrow interest: medical researcher}
\source{
General Social Survey https://gss.norc.org
}
\description{
medbstgm
}
\details{
Question Please look at Card B18. When making policy recommendations about genetically modified foods, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
A. Medical researchers
}
\section{Values}{

\itemize{
\item \verb{[1]} what is best for the country
\item \verb{[5]} own narrow interests
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3394/vshow}{GSS Data Explorer page for \code{medbstgm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab Total \cr
   2006 \tab 3574 \tab 311 \tab 269 \tab 185 \tab 64 \tab 59 \tab 48 \tab 4510 \cr
   Total \tab 3574 \tab 311 \tab 269 \tab 185 \tab 64 \tab 59 \tab 48 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgm}},
\code{\link{bizinfgm}},
\code{\link{eatgm}},
\code{\link{gmbiz}},
\code{\link{medagrgm}},
\code{\link{medinfgm}},
\code{\link{polbstgm}},
\code{\link{polinfgm}},
\code{\link{sciimp1}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
