% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{medcare2}
\alias{medcare2}
\title{Who r thinks should be next most responsible for cost of xs medical care}
\source{
General Social Survey https://gss.norc.org
}
\description{
medcare2
}
\details{
Question Who should be next most responsible?
}
\section{Values}{

\itemize{
\item \code{1} his-her family
\item \code{2} government
\item \code{3} insurance
\item \code{4} private charity
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2037/vshow}{GSS Data Explorer page for \code{medcare2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab his-her family \tab government \tab insurance \tab private charity \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2336 \tab 258 \tab 50 \tab 192 \tab 25 \tab 15 \tab 28 \tab 2904 \cr
   2006 \tab 4093 \tab 168 \tab 27 \tab 192 \tab 9 \tab 3 \tab 18 \tab 4510 \cr
   Total \tab 6429 \tab 426 \tab 77 \tab 384 \tab 34 \tab 18 \tab 46 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{mntlas1}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}},
\code{\link{spmentl}}
}
\concept{Mental Health}
\keyword{variable}
