% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{meddoc}
\alias{meddoc}
\title{X should go to general medical doctor for help}
\source{
General Social Survey https://gss.norc.org
}
\description{
meddoc
}
\details{
Question Should NAME do any of the following:
C. Go to a general medical doctor for help
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2013/vshow}{GSS Data Explorer page for \code{meddoc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 1019 \tab 355 \tab 43 \tab 27 \tab 2904 \cr
   2006 \tab 3073 \tab 1203 \tab 187 \tab 33 \tab 14 \tab 4510 \cr
   2018 \tab 1175 \tab 985 \tab 149 \tab 11 \tab 28 \tab 2348 \cr
   Total \tab 5708 \tab 3207 \tab 691 \tab 87 \tab 69 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dofirst}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}
}
\concept{Mental Health}
\keyword{variable}
