% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_580.R
\name{mediarel}
\alias{mediarel}
\title{Watch a religious program on television-radio?}
\source{
General Social Survey https://gss.norc.org
}
\description{
mediarel
}
\details{
Question During the last seven days did you do the following:
B. Watch a religious program on television or listen to a religious program on the radio?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/C/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5319/vshow}{GSS Data Explorer page for \code{mediarel}}.
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attreg}},
\code{\link{friday}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}

Other Religion: 
\code{\link{attreg}},
\code{\link{friday}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}
}
\concept{Religion}
\concept{Religious Attendance and Identity}
\keyword{variable}
