% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_375.R
\name{medinfgm}
\alias{medinfgm}
\title{Influence on restricting gmf by medical researcher's}
\source{
General Social Survey https://gss.norc.org
}
\description{
medinfgm
}
\details{
Question Some say that the government should restrict the sale of genetically modified foods. Others say there is no need for such restrictions. Please look at Card B17. How much influence should each of the following groups have in deciding whether to restrict the sale of genetically modified foods?
A. Medical researchers. (Would you say a great deal of influence, a fair amount, a little influence, or none at all?
}
\section{Values}{

\itemize{
\item \code{1} a great deal of influence
\item \code{2} a fair amount
\item \code{3} a little influence
\item \code{4} none at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3391/vshow}{GSS Data Explorer page for \code{medinfgm}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgm}},
\code{\link{bizinfgm}},
\code{\link{eatgm}},
\code{\link{gmbiz}},
\code{\link{medagrgm}},
\code{\link{medbstgm}},
\code{\link{polbstgm}},
\code{\link{polinfgm}},
\code{\link{sciimp1}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
