% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{meditate}
\alias{meditate}
\title{How often meditate}
\source{
General Social Survey https://gss.norc.org
}
\description{
meditate
}
\details{
Question Within your religious or spiritual tradition, how often do you meditate?
}
\section{Values}{

\itemize{
\item \code{1} mre thn once/day
\item \code{2} once a day
\item \code{3} a few times/week
\item \code{4} once a week
\item \code{5} few times/month
\item \code{6} once a month
\item \code{7} less once/mnth
\item \code{8} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2091/vshow}{GSS Data Explorer page for \code{meditate}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{privpray}}

Other Religion: 
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{privpray}}
}
\concept{Religion}
\concept{Religion and Culture}
\keyword{variable}
