% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_649.R
\name{medpay}
\alias{medpay}
\title{Couldn't get care because couldn't pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
medpay
}
\details{
Question During the past 12 months did it ever happen that you did not get the medical treatment you needed because... You could not pay for it?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} did not need medical treatment
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7827/vshow}{GSS Data Explorer page for \code{medpay}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab did not need medical treatment \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2381 \tab 150 \tab 842 \tab 165 \tab 1 \tab 5 \tab 3544 \cr
   Total \tab 2381 \tab 150 \tab 842 \tab 165 \tab 1 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
