% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{medsavtx}
\alias{medsavtx}
\title{Govt programs for mental health treatment will}
\source{
General Social Survey https://gss.norc.org
}
\description{
medsavtx
}
\details{
Question Now, I'd like to ask you for your opinions on several issues. For each question, please tell me if you strongly agree, agree somewhat, disagree somewhat or strongly disagree with the statement.
E. In the long-run, government programs that provide mental health treatment to children will save taxpayers money.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree somewhat
\item \verb{[3]} disagree somewhat
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2965/vshow}{GSS Data Explorer page for \code{medsavtx}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree somewhat \tab disagree somewhat \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 209 \tab 514 \tab 385 \tab 174 \tab 99 \tab 12 \tab 2765 \cr
   Total \tab 1372 \tab 209 \tab 514 \tab 385 \tab 174 \tab 99 \tab 12 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{mntlrel}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
