% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_333.R
\name{medunacc}
\alias{medunacc}
\title{Med for mental health problem often have side of}
\source{
General Social Survey https://gss.norc.org
}
\description{
medunacc
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors that are intended to help people who are having problems with their emotions, nerves or their mental health.
D. These medications often have unacceptable side effects.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2977/vshow}{GSS Data Explorer page for \code{medunacc}}.
}

\seealso{
Other Mental Health II: 
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{medaddct}},
\code{\link{medsymps}},
\code{\link{medweak}},
\code{\link{solveown}}
}
\concept{Mental Health II}
\keyword{variable}
