% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{meetf2f1}
\alias{meetf2f1}
\title{How often meet face to face with customers, clients, patients}
\source{
General Social Survey https://gss.norc.org
}
\description{
meetf2f1
}
\details{
Question How often do you meet face to face with customers, clients, patients, or others to whom you provide products or services?  Would you say….
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} once per year or less, or
\item \code{3} several times per year
\item \code{4} several times per month
\item \code{5} several times per week
\item \code{6} daily
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3890/vshow}{GSS Data Explorer page for \code{meetf2f1}}.
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}

Other Job: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}

Other Technology: 
\code{\link{autonojb}},
\code{\link{meetf2f2}}

Other Work: 
\code{\link{autonojb}},
\code{\link{meetf2f2}}
}
\concept{Global Economics}
\concept{Job}
\concept{Technology}
\concept{Work}
\keyword{variable}
