% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{meeting}
\alias{meeting}
\title{Web make r attend a political meeting}
\source{
General Social Survey https://gss.norc.org
}
\description{
meeting
}
\details{
Question Have you ever taken any of the following actions because of something you learned about or discussed on the Web?
C. Attended a political meeting or demonstration?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2444/vshow}{GSS Data Explorer page for \code{meeting}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 26 \tab 253 \tab 85 \tab 2817 \cr
   2002 \tab 2411 \tab 46 \tab 263 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 72 \tab 516 \tab 130 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{petition}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
