% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{mekdcare}
\alias{mekdcare}
\title{Men ought to do more childcare than they do now}
\source{
General Social Survey https://gss.norc.org
}
\description{
mekdcare
}
\details{
Question To what extent do you agree or disagree...?
B. Men ought to do a larger share of childcare than they do now.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4428/vshow}{GSS Data Explorer page for \code{mekdcare}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1594 \tab 294 \tab 534 \tab 251 \tab 74 \tab 17 \tab 1 \tab 2765 \cr
   Total \tab 1594 \tab 294 \tab 534 \tab 251 \tab 74 \tab 17 \tab 1 \tab 2765 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mehhwork}},
\code{\link{repairs1}},
\code{\link{shopfood}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
