% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{meltpot}
\alias{meltpot}
\title{R attitude toward mixing cultures in us}
\source{
General Social Survey https://gss.norc.org
}
\description{
meltpot
}
\details{
Question Some people say that it is better for America if different racial and ethnic groups maintain their distinct cultures. Others say that it is better if groups change so that they blend into the larger society as in the idea of a melting pot. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that racial and ethnic groups should maintain their distinct cultures, and a score of 7 as meaning that groups should change so that they blend into the larger society. What score between 1 and 7 comes closest to the way you feel? (CIRCLE ONE):
}
\section{Values}{

\itemize{
\item \code{1} maintain distinct cultures
\item \code{7} blend in to larger society
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1599/vshow}{GSS Data Explorer page for \code{meltpot}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{asngovt}},
\code{\link{bilinged}},
\code{\link{blkgovt}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{econfutr}},
\code{\link{econpast}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hspasn10}},
\code{\link{hspgovt}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{jobaff}},
\code{\link{letin}},
\code{\link{obrespct}},
\code{\link{owneth}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{united}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Multiculturalism}
\keyword{variable}
