% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_99.R
\name{memgrp3}
\alias{memgrp3}
\title{Person #3 is member of r's group}
\source{
General Social Survey https://gss.norc.org
}
\description{
memgrp3
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
G3. Third Person is Group Member
}
\section{Values}{

\itemize{
\item \verb{[1]} mentioned
\item \verb{[2]} not mentioned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/907/vshow}{GSS Data Explorer page for \code{memgrp3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab don't know \tab Total \cr
   1985 \tab 599 \tab 153 \tab 778 \tab 4 \tab - \tab 1534 \cr
   1987 \tab 735 \tab 149 \tab 910 \tab 25 \tab - \tab 1819 \cr
   2004 \tab 2240 \tab 57 \tab 467 \tab 47 \tab 1 \tab 2812 \cr
   Total \tab 3574 \tab 359 \tab 2155 \tab 76 \tab 1 \tab 6165 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp4}},
\code{\link{memgrp5}}
}
\concept{Social Networks}
\keyword{variable}
