% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memhobby}
\alias{memhobby}
\title{Membership in hobby club}
\source{
General Social Survey https://gss.norc.org
}
\description{
memhobby
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
I. Hobby or garden clubs
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/971/vshow}{GSS Data Explorer page for \code{memhobby}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1974 \tab - \tab 143 \tab 1319 \tab 22 \tab - \tab 1484 \cr
   1975 \tab - \tab 129 \tab 1327 \tab 32 \tab 2 \tab 1490 \cr
   1977 \tab - \tab 141 \tab 1375 \tab 14 \tab - \tab 1530 \cr
   1978 \tab - \tab 143 \tab 1374 \tab 15 \tab - \tab 1532 \cr
   1980 \tab - \tab 123 \tab 1320 \tab 24 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 155 \tab 1437 \tab 7 \tab - \tab 1599 \cr
   1984 \tab - \tab 129 \tab 1323 \tab 19 \tab 2 \tab 1473 \cr
   1986 \tab - \tab 130 \tab 1327 \tab 13 \tab - \tab 1470 \cr
   1987 \tab - \tab 147 \tab 1649 \tab 22 \tab 1 \tab 1819 \cr
   1988 \tab 484 \tab 103 \tab 882 \tab 12 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 88 \tab 917 \tab 30 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 104 \tab 788 \tab 7 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 110 \tab 898 \tab 8 \tab 1 \tab 1517 \cr
   1993 \tab 549 \tab 133 \tab 918 \tab 5 \tab 1 \tab 1606 \cr
   1994 \tab 2481 \tab 47 \tab 462 \tab 2 \tab - \tab 2992 \cr
   2004 \tab 1340 \tab 159 \tab 1308 \tab 5 \tab - \tab 2812 \cr
   Total \tab 6329 \tab 1984 \tab 18624 \tab 237 \tab 8 \tab 27182 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memgreek}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{memprof}},
\code{\link{memschl}}
}
\concept{Socio-political participation}
\keyword{variable}
