% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memnat}
\alias{memnat}
\title{Membership in nationality group}
\source{
General Social Survey https://gss.norc.org
}
\description{
memnat
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
K. Nationality groups
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/973/vshow}{GSS Data Explorer page for \code{memnat}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1974 \tab - \tab 52 \tab 1410 \tab 2 \tab 20 \tab 1484 \cr
   1975 \tab - \tab 37 \tab 1417 \tab 2 \tab 34 \tab 1490 \cr
   1977 \tab - \tab 52 \tab 1466 \tab - \tab 12 \tab 1530 \cr
   1978 \tab - \tab 39 \tab 1475 \tab - \tab 18 \tab 1532 \cr
   1980 \tab - \tab 37 \tab 1400 \tab 2 \tab 29 \tab 1468 \cr
   1983 \tab - \tab 57 \tab 1535 \tab - \tab 7 \tab 1599 \cr
   1984 \tab - \tab 49 \tab 1401 \tab 2 \tab 21 \tab 1473 \cr
   1986 \tab - \tab 68 \tab 1392 \tab - \tab 10 \tab 1470 \cr
   1987 \tab - \tab 45 \tab 1750 \tab 1 \tab 23 \tab 1819 \cr
   1988 \tab 484 \tab 22 \tab 963 \tab - \tab 12 \tab 1481 \cr
   1989 \tab 502 \tab 35 \tab 968 \tab - \tab 32 \tab 1537 \cr
   1990 \tab 473 \tab 41 \tab 849 \tab 1 \tab 8 \tab 1372 \cr
   1991 \tab 500 \tab 46 \tab 959 \tab 2 \tab 10 \tab 1517 \cr
   1993 \tab 549 \tab 50 \tab 998 \tab 3 \tab 6 \tab 1606 \cr
   1994 \tab 2481 \tab 18 \tab 491 \tab - \tab 2 \tab 2992 \cr
   2004 \tab 1340 \tab 39 \tab 1428 \tab - \tab 5 \tab 2812 \cr
   Total \tab 6329 \tab 687 \tab 19902 \tab 15 \tab 249 \tab 27182 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{memprof}},
\code{\link{memschl}}
}
\concept{Socio-political participation}
\keyword{variable}
