% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memnum}
\alias{memnum}
\title{Number of memberships}
\source{
General Social Survey https://gss.norc.org
}
\description{
memnum
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
Q. Any membership at all
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/979/vshow}{GSS Data Explorer page for \code{memnum}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 15 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab 14 \tab 16 \tab don't know \tab Total \cr
   1974 \tab - \tab 375 \tab 381 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 281 \tab 185 \tab 114 \tab 67 \tab 28 \tab 21 \tab 11 \tab 8 \tab 3 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 410 \tab 424 \tab 4 \tab - \tab - \tab - \tab - \tab 261 \tab 169 \tab 97 \tab 46 \tab 38 \tab 23 \tab 8 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab 1490 \cr
   1977 \tab - \tab 433 \tab 423 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 247 \tab 168 \tab 106 \tab 68 \tab 33 \tab 24 \tab 12 \tab 9 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 426 \tab 418 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 271 \tab 202 \tab 106 \tab 46 \tab 28 \tab 19 \tab 4 \tab 6 \tab 2 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 498 \tab 371 \tab 3 \tab - \tab - \tab - \tab - \tab 242 \tab 164 \tab 81 \tab 43 \tab 30 \tab 13 \tab 9 \tab 5 \tab 8 \tab - \tab 1 \tab - \tab 1468 \cr
   1983 \tab - \tab 435 \tab 409 \tab 1 \tab - \tab 2 \tab - \tab - \tab 300 \tab 182 \tab 119 \tab 58 \tab 52 \tab 18 \tab 17 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 474 \tab 347 \tab 8 \tab 3 \tab - \tab - \tab - \tab 256 \tab 158 \tab 89 \tab 53 \tab 36 \tab 19 \tab 12 \tab 6 \tab 11 \tab 1 \tab - \tab - \tab 1473 \cr
   1986 \tab - \tab 408 \tab 363 \tab - \tab 1 \tab - \tab - \tab - \tab 259 \tab 168 \tab 127 \tab 66 \tab 37 \tab 24 \tab 8 \tab 6 \tab 3 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 606 \tab 478 \tab 3 \tab - \tab 1 \tab - \tab - \tab 314 \tab 159 \tab 97 \tab 72 \tab 33 \tab 25 \tab 13 \tab 5 \tab 11 \tab - \tab 2 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 294 \tab 256 \tab - \tab - \tab - \tab - \tab - \tab 187 \tab 102 \tab 71 \tab 36 \tab 28 \tab 11 \tab 7 \tab 5 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 502 \tab 313 \tab 243 \tab - \tab - \tab - \tab - \tab - \tab 191 \tab 120 \tab 72 \tab 41 \tab 20 \tab 16 \tab 9 \tab 5 \tab 4 \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 286 \tab 218 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 154 \tab 86 \tab 69 \tab 39 \tab 25 \tab 8 \tab 5 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 500 \tab 329 \tab 242 \tab 2 \tab - \tab 1 \tab - \tab - \tab 182 \tab 107 \tab 73 \tab 37 \tab 22 \tab 10 \tab 7 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 549 \tab 298 \tab 255 \tab 1 \tab 1 \tab - \tab - \tab - \tab 183 \tab 124 \tab 86 \tab 50 \tab 25 \tab 15 \tab 10 \tab 5 \tab 4 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2481 \tab 146 \tab 125 \tab - \tab - \tab - \tab - \tab - \tab 82 \tab 67 \tab 34 \tab 27 \tab 12 \tab 9 \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2992 \cr
   2004 \tab 1340 \tab 555 \tab 329 \tab 1 \tab - \tab - \tab 1 \tab - \tab 209 \tab 141 \tab 102 \tab 52 \tab 41 \tab 22 \tab 11 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 2812 \cr
   Total \tab 6329 \tab 6286 \tab 5282 \tab 32 \tab 12 \tab 10 \tab 4 \tab 1 \tab 3619 \tab 2302 \tab 1443 \tab 801 \tab 488 \tab 277 \tab 149 \tab 78 \tab 62 \tab 2 \tab 4 \tab 1 \tab 27182 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memother}},
\code{\link{memprof}},
\code{\link{memschl}}
}
\concept{Socio-political participation}
\keyword{variable}
