% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{memsport}
\alias{memsport}
\title{Membership in sports club}
\source{
General Social Survey https://gss.norc.org
}
\description{
memsport
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
F. Sports groups
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/968/vshow}{GSS Data Explorer page for \code{memsport}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1974 \tab - \tab 262 \tab 1202 \tab 1 \tab 19 \tab 1484 \cr
   1975 \tab - \tab 278 \tab 1186 \tab 1 \tab 25 \tab 1490 \cr
   1977 \tab - \tab 283 \tab 1234 \tab - \tab 13 \tab 1530 \cr
   1978 \tab - \tab 297 \tab 1223 \tab - \tab 12 \tab 1532 \cr
   1980 \tab - \tab 249 \tab 1196 \tab 1 \tab 22 \tab 1468 \cr
   1983 \tab - \tab 339 \tab 1254 \tab - \tab 6 \tab 1599 \cr
   1984 \tab - \tab 311 \tab 1142 \tab 2 \tab 18 \tab 1473 \cr
   1986 \tab - \tab 307 \tab 1154 \tab - \tab 9 \tab 1470 \cr
   1987 \tab - \tab 327 \tab 1472 \tab 1 \tab 19 \tab 1819 \cr
   1988 \tab 484 \tab 194 \tab 795 \tab - \tab 8 \tab 1481 \cr
   1989 \tab 502 \tab 218 \tab 788 \tab - \tab 29 \tab 1537 \cr
   1990 \tab 473 \tab 173 \tab 721 \tab - \tab 5 \tab 1372 \cr
   1991 \tab 500 \tab 178 \tab 827 \tab 1 \tab 11 \tab 1517 \cr
   1993 \tab 549 \tab 196 \tab 855 \tab 1 \tab 5 \tab 1606 \cr
   1994 \tab 2481 \tab 110 \tab 399 \tab - \tab 2 \tab 2992 \cr
   2004 \tab 1340 \tab 246 \tab 1221 \tab - \tab 5 \tab 2812 \cr
   Total \tab 6329 \tab 3968 \tab 16669 \tab 8 \tab 208 \tab 27182 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{memfrat}},
\code{\link{mempolit}},
\code{\link{memserv}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}}
}
\concept{Socio-political participation}
\keyword{variable}
