% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{memvet}
\alias{memvet}
\title{Membership in veteran group}
\source{
General Social Survey https://gss.norc.org
}
\description{
memvet
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
C. Veterans' groups
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/965/vshow}{GSS Data Explorer page for \code{memvet}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1974 \tab - \tab 132 \tab 1332 \tab 20 \tab - \tab 1484 \cr
   1975 \tab - \tab 114 \tab 1350 \tab 25 \tab 1 \tab 1490 \cr
   1977 \tab - \tab 126 \tab 1392 \tab 12 \tab - \tab 1530 \cr
   1978 \tab - \tab 99 \tab 1416 \tab 16 \tab 1 \tab 1532 \cr
   1980 \tab - \tab 108 \tab 1333 \tab 25 \tab 2 \tab 1468 \cr
   1983 \tab - \tab 106 \tab 1485 \tab 8 \tab - \tab 1599 \cr
   1984 \tab - \tab 101 \tab 1350 \tab 20 \tab 2 \tab 1473 \cr
   1986 \tab - \tab 91 \tab 1370 \tab 9 \tab - \tab 1470 \cr
   1987 \tab - \tab 105 \tab 1692 \tab 21 \tab 1 \tab 1819 \cr
   1988 \tab 484 \tab 84 \tab 904 \tab 9 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 78 \tab 929 \tab 28 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 64 \tab 829 \tab 6 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 71 \tab 937 \tab 9 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 71 \tab 980 \tab 5 \tab 1 \tab 1606 \cr
   1994 \tab 2481 \tab 40 \tab 469 \tab 2 \tab - \tab 2992 \cr
   2004 \tab 1340 \tab 78 \tab 1389 \tab 5 \tab - \tab 2812 \cr
   Total \tab 6329 \tab 1468 \tab 19157 \tab 220 \tab 8 \tab 27182 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{memfrat}},
\code{\link{mempolit}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memyouth}}
}
\concept{Socio-political participation}
\keyword{variable}
