% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{memyouth}
\alias{memyouth}
\title{Membership in youth group}
\source{
General Social Survey https://gss.norc.org
}
\description{
memyouth
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
G. Youth groups
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1974, 1975, 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/969/vshow}{GSS Data Explorer page for \code{memyouth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1974 \tab - \tab 153 \tab 1311 \tab 20 \tab - \tab 1484 \cr
   1975 \tab - \tab 144 \tab 1318 \tab 28 \tab - \tab 1490 \cr
   1977 \tab - \tab 157 \tab 1361 \tab 12 \tab - \tab 1530 \cr
   1978 \tab - \tab 137 \tab 1378 \tab 16 \tab 1 \tab 1532 \cr
   1980 \tab - \tab 117 \tab 1325 \tab 25 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 168 \tab 1423 \tab 8 \tab - \tab 1599 \cr
   1984 \tab - \tab 138 \tab 1313 \tab 20 \tab 2 \tab 1473 \cr
   1986 \tab - \tab 155 \tab 1305 \tab 10 \tab - \tab 1470 \cr
   1987 \tab - \tab 152 \tab 1644 \tab 22 \tab 1 \tab 1819 \cr
   1988 \tab 484 \tab 107 \tab 877 \tab 13 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 96 \tab 906 \tab 33 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 94 \tab 798 \tab 7 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 87 \tab 919 \tab 10 \tab 1 \tab 1517 \cr
   1993 \tab 549 \tab 105 \tab 945 \tab 5 \tab 2 \tab 1606 \cr
   1994 \tab 2481 \tab 53 \tab 455 \tab 2 \tab 1 \tab 2992 \cr
   2004 \tab 1340 \tab 151 \tab 1315 \tab 5 \tab 1 \tab 2812 \cr
   Total \tab 6329 \tab 2014 \tab 18593 \tab 236 \tab 10 \tab 27182 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{memfrat}},
\code{\link{mempolit}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}}
}
\concept{Socio-political participation}
\keyword{variable}
