% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_180.R
\name{menhrt}
\alias{menhrt}
\title{Men hurt from traditional family}
\source{
General Social Survey https://gss.norc.org
}
\description{
menhrt
}
\details{
Question Who is hurt by this arrangement: men, women, children, everyone, or no one? PROBE:  Anyone else? UNTIL R SAYS NO.
A. Men
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1683/vshow}{GSS Data Explorer page for \code{menhrt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 207 \tab 803 \tab 84 \tab 366 \tab 2904 \cr
   Total \tab 1444 \tab 207 \tab 803 \tab 84 \tab 366 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{fekids1}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{womenhrt}}
}
\concept{Gender}
\keyword{variable}
