% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{mentlhos}
\alias{mentlhos}
\title{X should check into a mental hospital}
\source{
General Social Survey https://gss.norc.org
}
\description{
mentlhos
}
\details{
Question Should NAME do any of the following:
J. Check into a mental hospital.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2020/vshow}{GSS Data Explorer page for \code{mentlhos}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1460 \tab 441 \tab 836 \tab 132 \tab 35 \tab - \tab - \tab 2904 \cr
   2006 \tab 3073 \tab 465 \tab 879 \tab 76 \tab 17 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 469 \tab 622 \tab 55 \tab 27 \tab - \tab - \tab 2348 \cr
   2024 \tab 1693 \tab 675 \tab 806 \tab 89 \tab 12 \tab - \tab 34 \tab 3309 \cr
   Total \tab 7401 \tab 2050 \tab 3143 \tab 352 \tab 91 \tab 0 \tab 34 \tab 13071 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}

Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}
}
\concept{Mental Health}
\keyword{variable}
