% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{mentlill}
\alias{mentlill}
\title{How likely x is experiencing mental illness}
\source{
General Social Survey https://gss.norc.org
}
\description{
mentlill
}
\details{
Question In your opinion, how likely is it that NAME is experiencing (READ A-E) -- very likely, somewhat likely, not very likely, or not at all likely? (IN PART E, READ ONLY THE PHRASE FOLLOWING THE LETTER OF THE VIGNETTE)
C. A mental illness
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1996/vshow}{GSS Data Explorer page for \code{mentlill}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1460 \tab 285 \tab 460 \tab 395 \tab 216 \tab 64 \tab 24 \tab - \tab - \tab 2904 \cr
   2006 \tab 3073 \tab 355 \tab 504 \tab 372 \tab 147 \tab 44 \tab 15 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 278 \tab 414 \tab 289 \tab 140 \tab 24 \tab 28 \tab - \tab - \tab 2348 \cr
   2024 \tab 1691 \tab 385 \tab 650 \tab 351 \tab 157 \tab 39 \tab 11 \tab - \tab 25 \tab 3309 \cr
   Total \tab 7399 \tab 1303 \tab 2028 \tab 1407 \tab 660 \tab 171 \tab 78 \tab 0 \tab 25 \tab 13071 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}

Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
