% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{meovrwrk}
\alias{meovrwrk}
\title{Men hurt family when focus on work too much}
\source{
General Social Survey https://gss.norc.org
}
\description{
meovrwrk
}
\details{
Question And, do you agree or disagree . . .
C. Family life often suffers because men concentrate too much on their work.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1994, 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4404/vshow}{GSS Data Explorer page for \code{meovrwrk}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1994 \tab 1545 \tab 122 \tab 695 \tab 286 \tab 243 \tab 41 \tab 33 \tab 27 \tab - \tab 2992 \cr
   1996 \tab 1444 \tab 230 \tab 825 \tab 169 \tab 198 \tab 21 \tab 16 \tab 1 \tab - \tab 2904 \cr
   2000 \tab 940 \tab 209 \tab 877 \tab 331 \tab 361 \tab 34 \tab 43 \tab 22 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 99 \tab 415 \tab 108 \tab 264 \tab 16 \tab 6 \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 94 \tab 460 \tab 135 \tab 188 \tab 25 \tab 4 \tab - \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 208 \tab 945 \tab 304 \tab 477 \tab 43 \tab 14 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 143 \tab 653 \tab 161 \tab 310 \tab 50 \tab 12 \tab - \tab - \tab 2023 \cr
   2010 \tab 614 \tab 122 \tab 662 \tab 192 \tab 388 \tab 57 \tab 6 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 130 \tab 558 \tab 170 \tab 382 \tab 51 \tab 11 \tab - \tab - \tab 1974 \cr
   2014 \tab 863 \tab 176 \tab 702 \tab 234 \tab 479 \tab 76 \tab 7 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 171 \tab 819 \tab 257 \tab 536 \tab 96 \tab 9 \tab - \tab - \tab 2867 \cr
   2018 \tab 789 \tab 134 \tab 644 \tab 220 \tab 475 \tab 73 \tab 11 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 202 \tab 886 \tab 1001 \tab 487 \tab 138 \tab 1 \tab - \tab 2 \tab 4032 \cr
   2022 \tab 1168 \tab 201 \tab 885 \tab 618 \tab 537 \tab 117 \tab 15 \tab 1 \tab 2 \tab 3544 \cr
   2024 \tab 1126 \tab 195 \tab 787 \tab 611 \tab 481 \tab 89 \tab 19 \tab - \tab 1 \tab 3309 \cr
   Total \tab 18430 \tab 2436 \tab 10813 \tab 4797 \tab 5806 \tab 927 \tab 207 \tab 58 \tab 5 \tab 43479 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{balwkfam}},
\code{\link{children}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{earnmore}},
\code{\link{emoteoth}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{febear}},
\code{\link{fehire}},
\code{\link{feimp}},
\code{\link{fejobaff}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{flextime}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{haprom}},
\code{\link{hmemaker}},
\code{\link{livewith}},
\code{\link{manprof}},
\code{\link{marfree}},
\code{\link{mebear}},
\code{\link{men}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{parleave}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{rhmewrk}},
\code{\link{romance}},
\code{\link{rushed}},
\code{\link{selffrst}},
\code{\link{sexhar}},
\code{\link{sexpromo}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{twoincs}},
\code{\link{willwed1}},
\code{\link{willwed2}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{wrkclass}},
\code{\link{yourself}}

Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{buythngs}},
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{chldcare}},
\code{\link{clean1}},
\code{\link{cohabfst}},
\code{\link{cohabit}},
\code{\link{cohabok}},
\code{\link{cooking1}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{deckids}},
\code{\link{dinner}},
\code{\link{divbest}},
\code{\link{divhubby}},
\code{\link{divifkd1}},
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokd1}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{eldcost}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{evdiv}},
\code{\link{fambudgt}},
\code{\link{famsuffr}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{fejobind}},
\code{\link{femarry}},
\code{\link{fewknokd}},
\code{\link{fewrksup}},
\code{\link{fourkids}},
\code{\link{hapifwrk}},
\code{\link{happy7}},
\code{\link{hhclean1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}},
\code{\link{kidsocst}},
\code{\link{kidsuffr}},
\code{\link{kidtrble}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{mardiv}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{mawork14}},
\code{\link{mawrkwrm}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{memarry}},
\code{\link{mrmom}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{rfamlook}},
\code{\link{rhhwork}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{sexhar}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{singlpar}},
\code{\link{spborn}},
\code{\link{spevdiv}},
\code{\link{spfalook}},
\code{\link{sphhwork}},
\code{\link{splive}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{teensex1}},
\code{\link{threkids}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhm1}},
\code{\link{tiredhme}},
\code{\link{tiredwk1}},
\code{\link{tiredwrk}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{twokids}},
\code{\link{weekend}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}

Other Family: 
\code{\link{adoptkid}},
\code{\link{agewed}},
\code{\link{agewedcr}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{balwkfam}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{chldben}},
\code{\link{chlddth}},
\code{\link{chldhrt}},
\code{\link{clsclsex}},
\code{\link{cousins}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce}},
\code{\link{engda}},
\code{\link{engson}},
\code{\link{evdiv}},
\code{\link{evkid}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{famlife}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famvswk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{feovrwrk}},
\code{\link{flextime}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{inlawnum}},
\code{\link{inperson}},
\code{\link{jobvsfam}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidcall}},
\code{\link{kidlived}},
\code{\link{kidmost}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{kidvisit}},
\code{\link{letters}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{madeath}},
\code{\link{malive}},
\code{\link{malive1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{marital}},
\code{\link{marunion}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{meetings}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{mykids}},
\code{\link{niecenep}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{numcntct}},
\code{\link{numprobs}},
\code{\link{obtohelp}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{othkids}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{padeath}},
\code{\link{palive}},
\code{\link{palive1}},
\code{\link{parleave}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{pawrkgrw}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{posslq}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relcall}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relmarry}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{rfamlook}},
\code{\link{sibcall}},
\code{\link{sibdeath}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibvisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{spdeath}},
\code{\link{spfalook}},
\code{\link{spfam16}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{sppart}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{widowed}},
\code{\link{wkvsfam}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}}

Other Men: 
\code{\link{hubbywk1}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{mrmom}}

Other Women: 
\code{\link{children}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{earnmore}},
\code{\link{famsuffr}},
\code{\link{febear}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fecolop}},
\code{\link{fefam}},
\code{\link{fefirm}},
\code{\link{fegains}},
\code{\link{fehelp}},
\code{\link{fehire}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinflu}},
\code{\link{feinfo}},
\code{\link{fejobaff}},
\code{\link{fejobind}},
\code{\link{fejobop}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{fewknokd}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrksup}},
\code{\link{hapifwrk}},
\code{\link{hmemaker}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{kidsuffr}},
\code{\link{manprof}},
\code{\link{mawrkwrm}},
\code{\link{mebear}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{men}},
\code{\link{mrmom}},
\code{\link{opsex}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{salfergt}},
\code{\link{sexpromo}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{wrkbaby}},
\code{\link{wrkclass}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}},
\code{\link{yourself}}
}
\concept{Family}
\concept{Gender}
\concept{ISSP Family & Gender Roles}
\concept{Men}
\concept{Women}
\keyword{variable}
