% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{meserve}
\alias{meserve}
\title{Mandatory conscription for males?}
\source{
General Social Survey https://gss.norc.org
}
\description{
meserve
}
\details{
Question B. And how would you feel about such a program for all young men--Would you strongly favor it, probably favor it, probably oppose it, or strongly oppose it?
RECORD UNDER "B."
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} probably favor
\item \verb{[3]} probably oppose
\item \verb{[4]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/779/vshow}{GSS Data Explorer page for \code{meserve}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly favor \tab probably favor \tab probably oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 658 \tab 670 \tab 269 \tab 179 \tab 61 \tab 23 \tab 1860 \cr
   1984 \tab - \tab 497 \tab 546 \tab 244 \tab 134 \tab 32 \tab 20 \tab 1473 \cr
   Total \tab 0 \tab 1155 \tab 1216 \tab 513 \tab 313 \tab 93 \tab 43 \tab 3333 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\keyword{variable}
