% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhclsoth}
\alias{mhclsoth}
\title{Closeness to person with mental health problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhclsoth
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
A. I want you to think about the person with a mental health problem with whom you have had the most contact, would you say you were extremely close, very close, not very close, or not at all close to that person?
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely close
\item \verb{[2]} very close
\item \verb{[3]} not very close
\item \verb{[4]} not at all close
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3803/vshow}{GSS Data Explorer page for \code{mhclsoth}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab extremely close \tab very close \tab not very close \tab not at all close \tab no answer \tab Total \cr
   2006 \tab 3535 \tab 323 \tab 381 \tab 204 \tab 61 \tab 6 \tab 4510 \cr
   Total \tab 3535 \tab 323 \tab 381 \tab 204 \tab 61 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhcause}},
\code{\link{mhdisoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhseroth}},
\code{\link{mhtrtot2}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}
}
\concept{Mental Health IV}
\keyword{variable}
