% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhdisoth}
\alias{mhdisoth}
\title{How much distress this person caused r}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdisoth
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
E. How much distress did this person's mental health problem cause you?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} quite a bit
\item \verb{[3]} a little
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3807/vshow}{GSS Data Explorer page for \code{mhdisoth}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a great deal \tab quite a bit \tab a little \tab not at all \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2006 \tab 3535 \tab 148 \tab 204 \tab 340 \tab 275 \tab 1 \tab 7 \tab - \tab 4510 \cr
   Total \tab 3535 \tab 148 \tab 204 \tab 340 \tab 275 \tab 1 \tab 7 \tab 0 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhaccptd}},
\code{\link{mhbrain}},
\code{\link{mhcause}},
\code{\link{mhchrctr}},
\code{\link{mhclsoth}},
\code{\link{mhcreatv}},
\code{\link{mhdecslf}},
\code{\link{mhdoactv}},
\code{\link{mhdodiet}},
\code{\link{mhdodoc}},
\code{\link{mhdoexrc}},
\code{\link{mhdofam}},
\code{\link{mhdofrgt}},
\code{\link{mhdofrnd}},
\code{\link{mhdogrp}},
\code{\link{mhdoheal}},
\code{\link{mhdoherb}},
\code{\link{mhdohosp}},
\code{\link{mhdolife}},
\code{\link{mhdomhp}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdootc}},
\code{\link{mhdooth}},
\code{\link{mhdopray}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhdorx}},
\code{\link{mhexpoth}},
\code{\link{mhfamshm}},
\code{\link{mhfeelem}},
\code{\link{mhfriend}},
\code{\link{mhfright}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhgvtjob}},
\code{\link{mhhired}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlpoth}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhhrdtlk}},
\code{\link{mhhsptrt}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhkdcare}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhluck}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhmoney}},
\code{\link{mhneihbr}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhnotch}},
\code{\link{mhnotell}},
\code{\link{mhnotpol}},
\code{\link{mhoutsdr}},
\code{\link{mhphyscl}},
\code{\link{mhprdctv}},
\code{\link{mhproblm}},
\code{\link{mhraised}},
\code{\link{mhreloth}},
\code{\link{mhresoth}},
\code{\link{mhsecret}},
\code{\link{mhseedoc}},
\code{\link{mhseepub}},
\code{\link{mhseroth}},
\code{\link{mhslfshm}},
\code{\link{mhsocial}},
\code{\link{mhstress}},
\code{\link{mhsupwrk}},
\code{\link{mhsymp}},
\code{\link{mhtrtot2}},
\code{\link{mhtrtslf}},
\code{\link{mhtrusty}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}},
\code{\link{mhupdown}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}},
\code{\link{vigvermy}}
}
\concept{Mental Health IV}
\keyword{variable}
