% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_408.R
\name{mhdofrnd}
\alias{mhdofrnd}
\title{What x should do: talk to friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdofrnd
}
\details{
Question What do you think (NAME) should do about this situation?
FI INSTRUCTION: Do not read responses, code up to 3 responses.
B. Talk to friends?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3722/vshow}{GSS Data Explorer page for \code{mhdofrnd}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 138 \tab 1269 \tab 12 \tab 6 \tab 4510 \cr
   Total \tab 3085 \tab 138 \tab 1269 \tab 12 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdodoc}},
\code{\link{mhdofam}},
\code{\link{mhdoheal}},
\code{\link{mhdomhp}},
\code{\link{mhdootc}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhproblm}},
\code{\link{vigvermy}}
}
\concept{Mental Health IV}
\keyword{variable}
