% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_410.R
\name{mhdonone}
\alias{mhdonone}
\title{What x should do: nothing}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdonone
}
\details{
Question What do you think (NAME) should do about this situation?
FI INSTRUCTION: Do not read responses, code up to 3 responses.
U. Nothing?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3741/vshow}{GSS Data Explorer page for \code{mhdonone}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 11 \tab 1396 \tab 12 \tab 6 \tab 4510 \cr
   Total \tab 3085 \tab 11 \tab 1396 \tab 12 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhbrain}},
\code{\link{mhchrctr}},
\code{\link{mhdomove}},
\code{\link{mhdooth}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhluck}},
\code{\link{mhraised}},
\code{\link{mhstress}}
}
\concept{Mental Health IV}
\keyword{variable}
