% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_410.R
\name{mhdooth}
\alias{mhdooth}
\title{What x should do: other mentions}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdooth
}
\details{
Question What do you think (NAME) should do about this situation?
FI INSTRUCTION: Do not read responses, code up to 3 responses.
T. Other Mentions (Specify: __)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3740/vshow}{GSS Data Explorer page for \code{mhdooth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 98 \tab 1309 \tab 12 \tab 6 \tab 4510 \cr
   Total \tab 3085 \tab 98 \tab 1309 \tab 12 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhbrain}},
\code{\link{mhchrctr}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhluck}},
\code{\link{mhraised}},
\code{\link{mhstress}}
}
\concept{Mental Health IV}
\keyword{variable}
