% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_408.R
\name{mhdopsyc}
\alias{mhdopsyc}
\title{What x should do: go to psychiatrist}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdopsyc
}
\details{
Question What do you think (NAME) should do about this situation?
FI INSTRUCTION: Do not read responses, code up to 3 responses.
E. Go to a psychiatrist?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3725/vshow}{GSS Data Explorer page for \code{mhdopsyc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 359 \tab 1048 \tab 12 \tab 6 \tab 4510 \cr
   Total \tab 3085 \tab 359 \tab 1048 \tab 12 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdodoc}},
\code{\link{mhdofam}},
\code{\link{mhdofrnd}},
\code{\link{mhdoheal}},
\code{\link{mhdomhp}},
\code{\link{mhdootc}},
\code{\link{mhdorel}},
\code{\link{mhproblm}},
\code{\link{vigvermy}}
}
\concept{Mental Health IV}
\keyword{variable}
