% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_408.R
\name{mhdorel}
\alias{mhdorel}
\title{What x should do: talk to relic leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdorel
}
\details{
Question What do you think (NAME) should do about this situation?
FI INSTRUCTION: Do not read responses, code up to 3 responses.
C. Talk to a religious leader?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3723/vshow}{GSS Data Explorer page for \code{mhdorel}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdodoc}},
\code{\link{mhdofam}},
\code{\link{mhdofrnd}},
\code{\link{mhdoheal}},
\code{\link{mhdomhp}},
\code{\link{mhdootc}},
\code{\link{mhdopsyc}},
\code{\link{mhproblm}},
\code{\link{vigvermy}}

Other Mental Health: 
\code{\link{mhdodoc}},
\code{\link{mhdofam}},
\code{\link{mhdofrnd}},
\code{\link{mhdoheal}},
\code{\link{mhdomhp}},
\code{\link{mhdootc}},
\code{\link{mhdopsyc}},
\code{\link{mhproblm}},
\code{\link{vigvermy}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
