% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_409.R
\name{mhdorx}
\alias{mhdorx}
\title{What x should do: take prescription medication}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdorx
}
\details{
Question What do you think (NAME) should do about this situation?
FI INSTRUCTION: Do not read responses, code up to 3 responses.
I. Take prescription medication?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3729/vshow}{GSS Data Explorer page for \code{mhdorx}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdoactv}},
\code{\link{mhdodiet}},
\code{\link{mhdoexrc}},
\code{\link{mhdofrgt}},
\code{\link{mhdogrp}},
\code{\link{mhdoherb}},
\code{\link{mhdohosp}},
\code{\link{mhdolife}},
\code{\link{mhdopray}}
}
\concept{Mental Health IV}
\keyword{variable}
