% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{mhexpoth}
\alias{mhexpoth}
\title{Effect of experiences with this person}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhexpoth
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
G. Overall, would you say that your experiences with this person have…
}
\section{Values}{

\itemize{
\item \code{1} made you a better person
\item \code{2} made you a worse person, or
\item \code{3} had no effect on you as a person
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3809/vshow}{GSS Data Explorer page for \code{mhexpoth}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab made you a better person \tab made you a worse person, or \tab had no effect on you as a person \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3535 \tab 558 \tab 6 \tab 397 \tab 6 \tab 8 \tab 4510 \cr
   Total \tab 3535 \tab 558 \tab 6 \tab 397 \tab 6 \tab 8 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhfright}},
\code{\link{mhreloth}},
\code{\link{mhseepub}},
\code{\link{mhsymp}},
\code{\link{mhtrtslf}}
}
\concept{Mental Health IV}
\keyword{variable}
