% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_415.R
\name{mhfamshm}
\alias{mhfamshm}
\title{X family should be ashamed to receive govt assistance}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhfamshm
}
\details{
Question The next few questions deal with the government's responsibility to help people like (NAME). For each statement please tell me if you think the government definitely should, probably should, probably should not, or definitely should not be responsible for helping people with situations like (NAME).
G. How about (NAME)'s family? Should (NAME)'s family be ashamed if (NAME) receives governmental assistance?
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3794/vshow}{GSS Data Explorer page for \code{mhfamshm}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhhsptrt}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhseedoc}},
\code{\link{mhslfshm}}

Other Mental Health: 
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhhsptrt}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhseedoc}},
\code{\link{mhslfshm}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
