% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{mhfright}
\alias{mhfright}
\title{How frightening most people are in public places}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhfright
}
\details{
Question Do you frequently, occasionally, rarely, or never see someone who seems to have a serious mental health problem in a public place?
A. How frightening are most people that you see in public places who appear to have a serious mental health problem?  Are they…
}
\section{Values}{

\itemize{
\item \code{1} very frightening
\item \code{2} somewhat frightening
\item \code{3} not very frightening
\item \code{4} not at all frightening
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3812/vshow}{GSS Data Explorer page for \code{mhfright}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhexpoth}},
\code{\link{mhreloth}},
\code{\link{mhseepub}},
\code{\link{mhsymp}},
\code{\link{mhtrtslf}}

Other Mental Health: 
\code{\link{mhexpoth}},
\code{\link{mhreloth}},
\code{\link{mhseepub}},
\code{\link{mhsymp}},
\code{\link{mhtrtslf}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
