% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_410.R
\name{mhgenes}
\alias{mhgenes}
\title{X s situation caused by: genetic problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhgenes
}
\details{
Question Is it very likely, somewhat likely, not very likely, or not at all likely that (NAME)''s situation is caused by…
E. A genetic or inherited problem.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3746/vshow}{GSS Data Explorer page for \code{mhgenes}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 206 \tab 793 \tab 293 \tab 85 \tab 41 \tab 7 \tab 4510 \cr
   Total \tab 3085 \tab 206 \tab 793 \tab 293 \tab 85 \tab 41 \tab 7 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhbrain}},
\code{\link{mhchrctr}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdooth}},
\code{\link{mhgod}},
\code{\link{mhluck}},
\code{\link{mhraised}},
\code{\link{mhstress}}
}
\concept{Mental Health IV}
\keyword{variable}
