% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_414.R
\name{mhhlpfam}
\alias{mhhlpfam}
\title{How important: turn to family for help}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhhlpfam
}
\details{
Question Now I'm going to read you a list of things that (NAME) might Do to get help with (her/his) situation. For each statement  please tell me using a scale from 1 to 10, where 1 is not at all important and 10 is very important, how important you think each would be for (NAME) to use in order to get help with (his/her) situation.
A. Turn to family for help.
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[10]} very important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3782/vshow}{GSS Data Explorer page for \code{mhhlpfam}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab not at all important \tab very important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 15 \tab 791 \tab 47 \tab 22 \tab 17 \tab 88 \tab 60 \tab 95 \tab 172 \tab 101 \tab 8 \tab 9 \tab 4510 \cr
   Total \tab 3085 \tab 15 \tab 791 \tab 47 \tab 22 \tab 17 \tab 88 \tab 60 \tab 95 \tab 172 \tab 101 \tab 8 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhnotch}},
\code{\link{mhsecret}},
\code{\link{mhsupwrk}}
}
\concept{Mental Health IV}
\keyword{variable}
