% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_414.R
\name{mhhlpfrd}
\alias{mhhlpfrd}
\title{How important: turn to friends for help}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhhlpfrd
}
\details{
Question Now I'm going to read you a list of things that (NAME) might Do to get help with (her/his) situation. For each statement  please tell me using a scale from 1 to 10, where 1 is not at all important and 10 is very important, how important you think each would be for (NAME) to use in order to get help with (his/her) situation.
B. Turn to friends for help.
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{10} very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3783/vshow}{GSS Data Explorer page for \code{mhhlpfrd}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhnotch}},
\code{\link{mhsecret}},
\code{\link{mhsupwrk}}

Other Mental Health: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhnotch}},
\code{\link{mhsecret}},
\code{\link{mhsupwrk}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
