% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_414.R
\name{mhhlpgp}
\alias{mhhlpgp}
\title{How important: go to medical doctor for help}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhhlpgp
}
\details{
Question Now I'm going to read you a list of things that (NAME) might Do to get help with (her/his) situation. For each statement  please tell me using a scale from 1 to 10, where 1 is not at all important and 10 is very important, how important you think each would be for (NAME) to use in order to get help with (his/her) situation.
D. Go to a general medical doctor for help.
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{10} very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3785/vshow}{GSS Data Explorer page for \code{mhhlpgp}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab not at all important \tab very important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 22 \tab 799 \tab 58 \tab 20 \tab 19 \tab 78 \tab 45 \tab 90 \tab 142 \tab 137 \tab 6 \tab 9 \tab 4510 \cr
   Total \tab 3085 \tab 22 \tab 799 \tab 58 \tab 20 \tab 19 \tab 78 \tab 45 \tab 90 \tab 142 \tab 137 \tab 6 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhnotch}},
\code{\link{mhsecret}},
\code{\link{mhsupwrk}}
}
\concept{Mental Health IV}
\keyword{variable}
