% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_415.R
\name{mhmeds}
\alias{mhmeds}
\title{X should be forced to take prescribed medication by law}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhmeds
}
\details{
Question Some countries have laws that force people with situations like (NAME)'s into psychiatric treatment. Do you think that people like (NAME) definitely should be, probably should be, probably shouldn't be, or definitely shouldn't be forced by law to…
B. Take medication prescribed by a doctor?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should be
\item \verb{[2]} probably should be
\item \verb{[3]} probably should not be
\item \verb{[4]} definitely should not be
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3796/vshow}{GSS Data Explorer page for \code{mhmeds}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 232 \tab 362 \tab 349 \tab 441 \tab 31 \tab 10 \tab 4510 \cr
   Total \tab 3085 \tab 232 \tab 362 \tab 349 \tab 441 \tab 31 \tab 10 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhfamshm}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhhsptrt}},
\code{\link{mhmnthsp}},
\code{\link{mhseedoc}},
\code{\link{mhslfshm}}
}
\concept{Mental Health IV}
\keyword{variable}
