% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_412.R
\name{mhmoney}
\alias{mhmoney}
\title{How able is x to decide on managing own money}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhmoney
}
\details{
Question How able is (NAME) to make decisions about managing (his/her) own money?
}
\section{Values}{

\itemize{
\item \verb{[1]} very able
\item \verb{[2]} moderately able
\item \verb{[3]} not very able
\item \verb{[4]} not at all able
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3761/vshow}{GSS Data Explorer page for \code{mhmoney}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very able \tab moderately able \tab not very able \tab not at all able \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 366 \tab 456 \tab 420 \tab 119 \tab 54 \tab 10 \tab 4510 \cr
   Total \tab 3085 \tab 366 \tab 456 \tab 420 \tab 119 \tab 54 \tab 10 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdecslf}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhnotpol}},
\code{\link{mhoutsdr}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}}
}
\concept{Mental Health IV}
\keyword{variable}
