% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_414.R
\name{mhnotch}
\alias{mhnotch}
\title{People like x shouldn't be allowed to teach children.}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhnotch
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
T. People like (NAME) should not be allowed to teach children.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3781/vshow}{GSS Data Explorer page for \code{mhnotch}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhsecret}},
\code{\link{mhsupwrk}}

Other Mental Health: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhsecret}},
\code{\link{mhsupwrk}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
