% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_413.R
\name{mhnotell}
\alias{mhnotell}
\title{X should feel afraid to tell about his situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhnotell
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
Q. (NAME) should feel afraid to tell others about (his/her) situation.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3778/vshow}{GSS Data Explorer page for \code{mhnotell}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 13 \tab 169 \tab 924 \tab 295 \tab 15 \tab 9 \tab 4510 \cr
   Total \tab 3085 \tab 13 \tab 169 \tab 924 \tab 295 \tab 15 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhaccptd}},
\code{\link{mhcreatv}},
\code{\link{mhfeelem}},
\code{\link{mhhired}},
\code{\link{mhhrdtlk}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhprdctv}},
\code{\link{mhtrusty}}
}
\concept{Mental Health IV}
\keyword{variable}
