% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_412.R
\name{mhnotpol}
\alias{mhnotpol}
\title{People like x shouldn't be allowed to hold public office}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhnotpol
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
G. People like (NAME) should not be allowed to hold public office.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3768/vshow}{GSS Data Explorer page for \code{mhnotpol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 109 \tab 401 \tab 606 \tab 259 \tab 40 \tab 10 \tab 4510 \cr
   Total \tab 3085 \tab 109 \tab 401 \tab 606 \tab 259 \tab 40 \tab 10 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdecslf}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhmoney}},
\code{\link{mhoutsdr}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}}
}
\concept{Mental Health IV}
\keyword{variable}
