% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{mhothrel}
\alias{mhothrel}
\title{Result of mental health problem to relationship with him}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhothrel
}
\details{
Question As a result of the mental health problem, would you say that your relationship with this person…
}
\section{Values}{

\itemize{
\item \verb{[1]} became stronger
\item \verb{[2]} became worse
\item \verb{[3]} was unchanged
\item \verb{[4]} ended as a result of the problem
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3567/vshow}{GSS Data Explorer page for \code{mhothrel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab became stronger \tab became worse \tab was unchanged \tab ended as a result of the problem \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3595 \tab 237 \tab 125 \tab 457 \tab 77 \tab 6 \tab 13 \tab 4510 \cr
   Total \tab 3595 \tab 237 \tab 125 \tab 457 \tab 77 \tab 6 \tab 13 \tab 4510 \cr
}
}

\seealso{
Other Mental Health III: 
\code{\link{mhothyou}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
