% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{mhothyou}
\alias{mhothyou}
\title{How much distress his mental health problem caused r}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhothyou
}
\details{
Question How much distress did this person's mental health problem cause you…
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} quite a bit
\item \verb{[3]} a little
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3566/vshow}{GSS Data Explorer page for \code{mhothyou}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab quite a bit \tab a little \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3595 \tab 240 \tab 179 \tab 281 \tab 200 \tab 1 \tab 14 \tab 4510 \cr
   Total \tab 3595 \tab 240 \tab 179 \tab 281 \tab 200 \tab 1 \tab 14 \tab 4510 \cr
}
}

\seealso{
Other Mental Health III: 
\code{\link{mhothrel}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
