% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_412.R
\name{mhoutsdr}
\alias{mhoutsdr}
\title{Getting treatment would make x outsider in community}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhoutsdr
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
A. Getting treatment would make (NAME) an outsider in (his/her) community.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3762/vshow}{GSS Data Explorer page for \code{mhoutsdr}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdecslf}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhmoney}},
\code{\link{mhnotpol}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}}
}
\concept{Mental Health IV}
\keyword{variable}
