% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_221.R
\name{mhp1r2}
\alias{mhp1r2}
\title{Relation to person 1 (2nd mentioned)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhp1r2
}
\details{
Question (If present) Second relationship with first person.
}
\section{Values}{

\itemize{
\item \verb{[1]} spouse/partner (current/ex)
\item \verb{[2]} parent
\item \verb{[3]} sibling
\item \verb{[4]} child
\item \verb{[5]} grandparent
\item \verb{[6]} grandchild
\item \verb{[7]} other family
\item \verb{[8]} coworker
\item \verb{[9]} neighbor
\item \verb{[10]} friend
\item \verb{[11]} member of religious community
\item \verb{[12]} acquaintance
\item \verb{[13]} student/classmate
\item \verb{[14]} relative of a friend
\item \verb{[15]} friend of a relative
\item \verb{[16]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7101/vshow}{GSS Data Explorer page for \code{mhp1r2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab friend \tab acquaintance \tab student/classmate \tab relative of a friend \tab other \tab no answer \tab Total \cr
   2018 \tab 1751 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 589 \tab 2348 \cr
   Total \tab 1751 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 589 \tab 2348 \cr
}
}

\keyword{variable}
