% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_221.R
\name{mhp2r1}
\alias{mhp2r1}
\title{Relation to person 2 (1st mentioned)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhp2r1
}
\details{
Question First relationship with second person.
}
\section{Values}{

\itemize{
\item \verb{[1]} spouse/partner (current/ex)
\item \verb{[2]} parent
\item \verb{[3]} sibling
\item \verb{[4]} child
\item \verb{[5]} grandparent
\item \verb{[6]} grandchild
\item \verb{[7]} other family
\item \verb{[8]} coworker
\item \verb{[9]} neighbor
\item \verb{[10]} friend
\item \verb{[11]} member of religious community
\item \verb{[12]} acquaintance
\item \verb{[13]} student/classmate
\item \verb{[14]} relative of a friend
\item \verb{[15]} friend of a relative
\item \verb{[16]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7102/vshow}{GSS Data Explorer page for \code{mhp2r1}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab spouse/partner (current/ex) \tab parent \tab sibling \tab child \tab grandparent \tab grandchild \tab other family \tab coworker \tab neighbor \tab friend \tab member of religious community \tab acquaintance \tab student/classmate \tab relative of a friend \tab friend of a relative \tab other \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1982 \tab 17 \tab 27 \tab 45 \tab 21 \tab 11 \tab 3 \tab 81 \tab 8 \tab 9 \tab 117 \tab 2 \tab 5 \tab 1 \tab 3 \tab 2 \tab 4 \tab 2 \tab 8 \tab 2348 \cr
   Total \tab 1982 \tab 17 \tab 27 \tab 45 \tab 21 \tab 11 \tab 3 \tab 81 \tab 8 \tab 9 \tab 117 \tab 2 \tab 5 \tab 1 \tab 3 \tab 2 \tab 4 \tab 2 \tab 8 \tab 2348 \cr
}
}

\keyword{variable}
