% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_221.R
\name{mhp3r2}
\alias{mhp3r2}
\title{Relation to person 3 (2nd mentioned)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhp3r2
}
\details{
Question (If present) Second relationship with third person.
}
\section{Values}{

\itemize{
\item \code{1} spouse/partner (current/ex)
\item \code{2} parent
\item \code{3} sibling
\item \code{4} child
\item \code{5} grandparent
\item \code{6} grandchild
\item \code{7} other family
\item \code{8} coworker
\item \code{9} neighbor
\item \code{10} friend
\item \code{11} member of religious community
\item \code{12} acquaintance
\item \code{13} student/classmate
\item \code{14} relative of a friend
\item \code{15} friend of a relative
\item \code{16} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7105/vshow}{GSS Data Explorer page for \code{mhp3r2}}.
}

\keyword{variable}
