% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_222.R
\name{mhp5r1}
\alias{mhp5r1}
\title{Relation to person 5 (1st mentioned)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhp5r1
}
\details{
Question First relationship with fifth person.
}
\section{Values}{

\itemize{
\item \code{1} spouse/partner (current/ex)
\item \code{2} parent
\item \code{3} sibling
\item \code{4} child
\item \code{5} grandparent
\item \code{6} grandchild
\item \code{7} other family
\item \code{8} coworker
\item \code{9} neighbor
\item \code{10} friend
\item \code{11} member of religious community
\item \code{12} acquaintance
\item \code{13} student/classmate
\item \code{14} relative of a friend
\item \code{15} friend of a relative
\item \code{16} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7108/vshow}{GSS Data Explorer page for \code{mhp5r1}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab spouse/partner (current/ex) \tab parent \tab sibling \tab child \tab grandparent \tab grandchild \tab other family \tab coworker \tab neighbor \tab friend \tab member of religious community \tab acquaintance \tab student/classmate \tab friend of a relative \tab other \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2245 \tab 1 \tab 1 \tab 10 \tab 1 \tab 1 \tab 2 \tab 30 \tab 4 \tab 2 \tab 36 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2348 \cr
   Total \tab 2245 \tab 1 \tab 1 \tab 10 \tab 1 \tab 1 \tab 2 \tab 30 \tab 4 \tab 2 \tab 36 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2348 \cr
}
}

\keyword{variable}
