% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_408.R
\name{mhproblm}
\alias{mhproblm}
\title{How serious x s situation is}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhproblm
}
\details{
Question Please remember, there are no right-or-wrong answers to these questions. Please think about the person I just described when answering these questions. First, how serious would you consider (NAME)'s situation to be. would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} very serious
\item \verb{[2]} moderately serious
\item \verb{[3]} not very serious
\item \verb{[4]} not at all seriou
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3720/vshow}{GSS Data Explorer page for \code{mhproblm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very serious \tab moderately serious \tab not very serious \tab not at all seriou \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 899 \tab 444 \tab 60 \tab 10 \tab 6 \tab 6 \tab 4510 \cr
   Total \tab 3085 \tab 899 \tab 444 \tab 60 \tab 10 \tab 6 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdodoc}},
\code{\link{mhdofam}},
\code{\link{mhdofrnd}},
\code{\link{mhdoheal}},
\code{\link{mhdomhp}},
\code{\link{mhdootc}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{vigvermy}}
}
\concept{Mental Health IV}
\keyword{variable}
