% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{mhreloth}
\alias{mhreloth}
\title{Relationship with this person as a result of mental health problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhreloth
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
F. As a result of the mental health problem, would you say that your relationship with this person…
}
\section{Values}{

\itemize{
\item \code{1} became stronger
\item \code{2} became worse
\item \code{3} was unchanged
\item \code{4} ended as a result of the problem
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3808/vshow}{GSS Data Explorer page for \code{mhreloth}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab became stronger \tab became worse \tab was unchanged \tab ended as a result of the problem \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3535 \tab 292 \tab 74 \tab 530 \tab 66 \tab 5 \tab 8 \tab 4510 \cr
   Total \tab 3535 \tab 292 \tab 74 \tab 530 \tab 66 \tab 5 \tab 8 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhexpoth}},
\code{\link{mhfright}},
\code{\link{mhseepub}},
\code{\link{mhsymp}},
\code{\link{mhtrtslf}}
}
\concept{Mental Health IV}
\keyword{variable}
