% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhresoth}
\alias{mhresoth}
\title{How much responsibility r had for this person}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhresoth
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
D. How much responsibility did you have for making day-to-day decisions for this person? Would you say…
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} quite a bit
\item \code{3} a little
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3806/vshow}{GSS Data Explorer page for \code{mhresoth}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a great deal \tab quite a bit \tab a little \tab not at all \tab no answer \tab Total \cr
   2006 \tab 3535 \tab 98 \tab 84 \tab 213 \tab 573 \tab 7 \tab 4510 \cr
   Total \tab 3535 \tab 98 \tab 84 \tab 213 \tab 573 \tab 7 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhcause}},
\code{\link{mhclsoth}},
\code{\link{mhdisoth}},
\code{\link{mhhlpoth}},
\code{\link{mhseroth}},
\code{\link{mhtrtot2}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}
}
\concept{Mental Health IV}
\keyword{variable}
