% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_415.R
\name{mhseedoc}
\alias{mhseedoc}
\title{X should be forced to be examined at a clinic by law}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhseedoc
}
\details{
Question Some countries have laws that force people with situations like (NAME)'s into psychiatric treatment. Do you think that people like (NAME) definitely should be, probably should be, probably shouldn't be, or definitely shouldn't be forced by law to…
A. To be examined at a clinic or by a doctor?
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3795/vshow}{GSS Data Explorer page for \code{mhseedoc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 191 \tab 308 \tab 404 \tab 485 \tab 27 \tab 10 \tab 4510 \cr
   Total \tab 3085 \tab 191 \tab 308 \tab 404 \tab 485 \tab 27 \tab 10 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhfamshm}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhhsptrt}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhslfshm}}
}
\concept{Mental Health IV}
\keyword{variable}
