% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{mhseepub}
\alias{mhseepub}
\title{How often r sees someone with mental health problem in public places}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhseepub
}
\details{
Question Do you frequently, occasionally, rarely, or never see someone who seems to have a serious mental health problem in a public place?
}
\section{Values}{

\itemize{
\item \verb{[1]} frequently
\item \verb{[2]} occasionally
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3811/vshow}{GSS Data Explorer page for \code{mhseepub}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab frequently \tab occasionally \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 223 \tab 474 \tab 465 \tab 253 \tab 5 \tab 5 \tab 4510 \cr
   Total \tab 3085 \tab 223 \tab 474 \tab 465 \tab 253 \tab 5 \tab 5 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhexpoth}},
\code{\link{mhfright}},
\code{\link{mhreloth}},
\code{\link{mhsymp}},
\code{\link{mhtrtslf}}
}
\concept{Mental Health IV}
\keyword{variable}
