% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_414.R
\name{mhsupwrk}
\alias{mhsupwrk}
\title{People like x shouldn't be allowed to supervise others at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhsupwrk
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
S. People like (NAME) should not be allowed to supervise others at work.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3780/vshow}{GSS Data Explorer page for \code{mhsupwrk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 45 \tab 419 \tab 687 \tab 208 \tab 56 \tab 10 \tab 4510 \cr
   Total \tab 3085 \tab 45 \tab 419 \tab 687 \tab 208 \tab 56 \tab 10 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhnotch}},
\code{\link{mhsecret}}
}
\concept{Mental Health IV}
\keyword{variable}
