% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{mhsymp}
\alias{mhsymp}
\title{How much r feels sympathy for ppl with mental hlth problem in public places}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhsymp
}
\details{
Question Do you frequently, occasionally, rarely, or never see someone who seems to have a serious mental health problem in a public place?
B. Would you say that you feel no sympathy, a little sympathy, quite a bit of sympathy, or a great deal of sympathy, For people that seem to have a serious mental health problem that you see in public places?
}
\section{Values}{

\itemize{
\item \code{1} no sympathy at all
\item \code{2} a little sympathy
\item \code{3} quite a bit of sympathy
\item \code{4} a great deal of sympathy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3813/vshow}{GSS Data Explorer page for \code{mhsymp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no sympathy at all \tab a little sympathy \tab quite a bit of sympathy \tab a great deal of sympathy \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3343 \tab 13 \tab 295 \tab 469 \tab 379 \tab 6 \tab 5 \tab 4510 \cr
   Total \tab 3343 \tab 13 \tab 295 \tab 469 \tab 379 \tab 6 \tab 5 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhexpoth}},
\code{\link{mhfright}},
\code{\link{mhreloth}},
\code{\link{mhseepub}},
\code{\link{mhtrtslf}}
}
\concept{Mental Health IV}
\keyword{variable}
