% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_223.R
\name{mhtreat1}
\alias{mhtreat1}
\title{Is person 1 currently receiving treatment for their mental health problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtreat1
}
\details{
Question Is (NAME 1) currently receiving treatment for their mental health problem?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} unsure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7120/vshow}{GSS Data Explorer page for \code{mhtreat1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab unsure \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1751 \tab 325 \tab 170 \tab 100 \tab 1 \tab 1 \tab 2348 \cr
   Total \tab 1751 \tab 325 \tab 170 \tab 100 \tab 1 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
