% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_223.R
\name{mhtreat4}
\alias{mhtreat4}
\title{Is person 4 currently receiving treatment for their mental health problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtreat4
}
\details{
Question Is (NAME 4) currently receiving treatment for their mental health problem?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} unsure
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7123/vshow}{GSS Data Explorer page for \code{mhtreat4}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab unsure \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2190 \tab 70 \tab 51 \tab 26 \tab 5 \tab 6 \tab 2348 \cr
   Total \tab 2190 \tab 70 \tab 51 \tab 26 \tab 5 \tab 6 \tab 2348 \cr
}
}

\keyword{variable}
