% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_226.R
\name{mhtreatd}
\alias{mhtreatd}
\title{R ever treated for mh problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtreatd
}
\details{
Question Did you get treatment for your mental health problem?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7148/vshow}{GSS Data Explorer page for \code{mhtreatd}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2018 \tab 2116 \tab 187 \tab 10 \tab 35 \tab - \tab - \tab 2348 \cr
   2024 \tab 2570 \tab 683 \tab 53 \tab 2 \tab - \tab 1 \tab 3309 \cr
   Total \tab 4686 \tab 870 \tab 63 \tab 37 \tab 0 \tab 1 \tab 5657 \cr
}
}

\keyword{variable}
