% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhtrtot2}
\alias{mhtrtot2}
\title{R knows someone who received treatment for mental health}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtrtot2
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3802/vshow}{GSS Data Explorer page for \code{mhtrtot2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2006 \tab 3085 \tab 969 \tab 448 \tab 2 \tab 6 \tab - \tab - \tab 4510 \cr
   2024 \tab 1693 \tab 1098 \tab 477 \tab 13 \tab 9 \tab - \tab 19 \tab 3309 \cr
   Total \tab 4778 \tab 2067 \tab 925 \tab 15 \tab 15 \tab 0 \tab 19 \tab 7819 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhcause}},
\code{\link{mhclsoth}},
\code{\link{mhdisoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhseroth}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}

Other Mental Health: 
\code{\link{mhcause}},
\code{\link{mhclsoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhseroth}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
