% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{mhtrtoth}
\alias{mhtrtoth}
\title{R personally has known someone who received treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtrtoth
}
\details{
Question Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3565/vshow}{GSS Data Explorer page for \code{mhtrtoth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3073 \tab 902 \tab 511 \tab 11 \tab 13 \tab 4510 \cr
   Total \tab 3073 \tab 902 \tab 511 \tab 11 \tab 13 \tab 4510 \cr
}
}

\seealso{
Other Mental Health III: 
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
