% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_412.R
\name{mhunsure}
\alias{mhunsure}
\title{People like x are unpredictable}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhunsure
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
E. People like (NAME) are unpredictable.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
   B/C \tab 2018 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3766/vshow}{GSS Data Explorer page for \code{mhunsure}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2006 \tab 3085 \tab 96 \tab 612 \tab 508 \tab 170 \tab 30 \tab 9 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 251 \tab 498 \tab 281 \tab 79 \tab 35 \tab 29 \tab - \tab - \tab 2348 \cr
   2024 \tab 1692 \tab 266 \tab 757 \tab 435 \tab 85 \tab 38 \tab 13 \tab - \tab 23 \tab 3309 \cr
   Total \tab 5952 \tab 613 \tab 1867 \tab 1224 \tab 334 \tab 103 \tab 51 \tab 0 \tab 23 \tab 10167 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdecslf}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhmoney}},
\code{\link{mhnotpol}},
\code{\link{mhoutsdr}},
\code{\link{mhuneasy}}

Other Mental Health: 
\code{\link{mhdecslf}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhmoney}},
\code{\link{mhnotpol}},
\code{\link{mhoutsdr}},
\code{\link{mhuneasy}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
