% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_411.R
\name{mhupdown}
\alias{mhupdown}
\title{X is experiencing normal ups-and-downs of life}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhupdown
}
\details{
Question Is it very likely, somewhat likely, not very likely, or not at all likely that (NAME)''s situation is caused by…
H. Is it very likely, somewhat likely, not very likely, or not at all likely that (NAME) is experiencing the normal ups-and-downs of life.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3749/vshow}{GSS Data Explorer page for \code{mhupdown}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhfriend}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhkdcare}},
\code{\link{mhneihbr}},
\code{\link{mhphyscl}},
\code{\link{mhsocial}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}}
}
\concept{Mental Health IV}
\keyword{variable}
