% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhviooth}
\alias{mhviooth}
\title{How likely x violent toward other people}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhviooth
}
\details{
Question A. How likely is it (NAME) would do something violent or harmful toward other people. Is it…
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3799/vshow}{GSS Data Explorer page for \code{mhviooth}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 82 \tab 357 \tab 600 \tab 278 \tab 97 \tab 11 \tab 4510 \cr
   Total \tab 3085 \tab 82 \tab 357 \tab 600 \tab 278 \tab 97 \tab 11 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhcause}},
\code{\link{mhclsoth}},
\code{\link{mhdisoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhseroth}},
\code{\link{mhtrtot2}},
\code{\link{mhvioslf}}
}
\concept{Mental Health IV}
\keyword{variable}
