% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_411.R
\name{mhwedrel}
\alias{mhwedrel}
\title{R would have x marry someone related to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhwedrel
}
\details{
Question Would you be definitely willing, probably willing, probably unwilling, or definitely unwilling…
F. To have (NAME) marry someone related to you?
}
\section{Values}{

\itemize{
\item \code{1} definitely willing
\item \code{2} probably willing
\item \code{3} probably unwilling
\item \code{4} definitely unwilling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3757/vshow}{GSS Data Explorer page for \code{mhwedrel}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhfriend}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhkdcare}},
\code{\link{mhneihbr}},
\code{\link{mhphyscl}},
\code{\link{mhsocial}},
\code{\link{mhupdown}},
\code{\link{mhwkwith}}

Other Mental Health: 
\code{\link{mhfriend}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhkdcare}},
\code{\link{mhneihbr}},
\code{\link{mhphyscl}},
\code{\link{mhsocial}},
\code{\link{mhupdown}},
\code{\link{mhwkwith}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
