% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_411.R
\name{mhwkwith}
\alias{mhwkwith}
\title{R would work closely with x on a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhwkwith
}
\details{
Question Would you be definitely willing, probably willing, probably unwilling, or definitely unwilling…
E. To work closely with (NAME) on a job?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely willing
\item \verb{[2]} probably willing
\item \verb{[3]} probably unwilling
\item \verb{[4]} definitely unwilling
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3756/vshow}{GSS Data Explorer page for \code{mhwkwith}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 310 \tab 601 \tab 342 \tab 142 \tab 24 \tab 6 \tab 4510 \cr
   Total \tab 3085 \tab 310 \tab 601 \tab 342 \tab 142 \tab 24 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhfriend}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhkdcare}},
\code{\link{mhneihbr}},
\code{\link{mhphyscl}},
\code{\link{mhsocial}},
\code{\link{mhupdown}},
\code{\link{mhwedrel}}
}
\concept{Mental Health IV}
\keyword{variable}
