% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{migage}
\alias{migage}
\title{Age when r came to u.s.}
\source{
General Social Survey https://gss.norc.org
}
\description{
migage
}
\details{
Question Constructed from variables AGE, YEARSUSA, and MNTHSUSA
NOTE: Please see the 2024 Codebook for more information on this variable. Please be aware that a programming error resulted in missing information for YEARSUSA and MNTHSUSA on Ballot 3 in GSS years 2018-2024.
}
\section{Values}{

\itemize{
\item \verb{[1]} migrated before age 5
\item \verb{[2]} migrated between ages 6-17
\item \verb{[3]} migrated between ages 18-29
\item \verb{[4]} migrated between ages 30-39
\item \verb{[5]} migrated age 40 or older
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8284/vshow}{GSS Data Explorer page for \code{migage}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab migrated before age 5 \tab migrated between ages 6-17 \tab migrated between ages 18-29 \tab migrated between ages 30-39 \tab migrated age 40 or older \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2018 \tab - \tab 46 \tab 57 \tab 117 \tab 46 \tab 33 \tab 2040 \tab 9 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 74 \tab 91 \tab 130 \tab 79 \tab 29 \tab 3288 \tab 333 \tab 1 \tab 7 \tab 4032 \cr
   2022 \tab - \tab 46 \tab 73 \tab 130 \tab 57 \tab 30 \tab 2994 \tab 211 \tab 2 \tab 1 \tab 3544 \cr
   2024 \tab - \tab 40 \tab 54 \tab 141 \tab 66 \tab 31 \tab 2870 \tab 101 \tab 1 \tab 5 \tab 3309 \cr
   Total \tab 0 \tab 206 \tab 275 \tab 518 \tab 248 \tab 123 \tab 11192 \tab 654 \tab 4 \tab 13 \tab 13233 \cr
}
}

\concept{Core}
\concept{Nativity}
\keyword{variable}
