% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{mighist}
\alias{mighist}
\title{When r came to u.s.}
\source{
General Social Survey https://gss.norc.org
}
\description{
mighist
}
\details{
Question Constructed from variables YEARSUSA and MNTHSUSA
NOTE: Please see the 2024 Codebook for more information on this variable. Please be aware that a programming error resulted in missing information for YEARSUSA and MNTHSUSA on Ballot 3 in GSS years 2018-2024.
}
\section{Values}{

\itemize{
\item \verb{[1]} migrated to the u.s. between 2012 and today
\item \verb{[2]} migrated to the u.s. between 2002 and 2011
\item \verb{[3]} migrated to the u.s. between 1996 and 2001
\item \verb{[4]} migrated to the u.s. between 1982 and 1995
\item \verb{[5]} migrated before 1982
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8285/vshow}{GSS Data Explorer page for \code{mighist}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab migrated to the u.s. between 2012 and today \tab migrated to the u.s. between 2002 and 2011 \tab migrated to the u.s. between 1996 and 2001 \tab migrated to the u.s. between 1982 and 1995 \tab migrated before 1982 \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2018 \tab - \tab 39 \tab 67 \tab 55 \tab 65 \tab 75 \tab 2045 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 96 \tab 91 \tab 58 \tab 91 \tab 111 \tab 3566 \tab - \tab 1 \tab 18 \tab 4032 \cr
   2022 \tab - \tab 104 \tab 74 \tab 51 \tab 66 \tab 61 \tab 3176 \tab 6 \tab 2 \tab 4 \tab 3544 \cr
   2024 \tab - \tab 135 \tab 56 \tab 46 \tab 64 \tab 42 \tab 2952 \tab 1 \tab 1 \tab 12 \tab 3309 \cr
   Total \tab 0 \tab 374 \tab 288 \tab 210 \tab 286 \tab 289 \tab 11739 \tab 9 \tab 4 \tab 34 \tab 13233 \cr
}
}

\seealso{
Other Core: 
\code{\link{conorelgsp}},
\code{\link{family16sex}},
\code{\link{femsee1}},
\code{\link{mascsee1}},
\code{\link{racdif5x}},
\code{\link{sexnow2}}
}
\concept{Core}
\concept{Nativity}
\keyword{variable}
