% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{migrpoor}
\alias{migrpoor}
\title{People from poor countries should be allowed to work in wealthy countries.}
\source{
General Social Survey https://gss.norc.org
}
\description{
migrpoor
}
\details{
Question People from poor countries should be allowed to work in wealthy countries.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7510/vshow}{GSS Data Explorer page for \code{migrpoor}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 280 \tab 770 \tab 498 \tab 145 \tab 49 \tab 91 \tab 156 \tab 14 \tab 4032 \cr
   Total \tab 2029 \tab 280 \tab 770 \tab 498 \tab 145 \tab 49 \tab 91 \tab 156 \tab 14 \tab 4032 \cr
}
}

\keyword{variable}
