% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{milokfe}
\alias{milokfe}
\title{Mil service good experience for women}
\source{
General Social Survey https://gss.norc.org
}
\description{
milokfe
}
\details{
Question B. How about for most young women, do you think military service is definitely a good experience, probably a good experience, probably not a good experience, or definitely not a good experience for them?
}
\section{Values}{

\itemize{
\item \code{1} definitely good
\item \code{2} probably good
\item \code{3} prob not good
\item \code{4} def not good
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/819/vshow}{GSS Data Explorer page for \code{milokfe}}.
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}
}
\concept{Core}
\keyword{variable}
