% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{milpay}
\alias{milpay}
\title{Pay and benefits for military personnel}
\source{
General Social Survey https://gss.norc.org
}
\description{
milpay
}
\details{
Question As you know, this country stopped the military draft in 1972. Since that time we have relied on volunteers. Now I'd like to ask you a few questions about our armed forces.
B. How do you feel about the pay and benefits that people in the armed forces receive? Would you say the pay  and benefits should be larger than they are now, should they be smaller, or are they about the right amount?
}
\section{Values}{

\itemize{
\item \verb{[1]} should be larger
\item \verb{[2]} about right
\item \verb{[3]} should be smaller
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/782/vshow}{GSS Data Explorer page for \code{milpay}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab should be larger \tab about right \tab should be smaller \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 683 \tab 897 \tab 33 \tab 227 \tab 20 \tab 1860 \cr
   1983 \tab - \tab 434 \tab 925 \tab 33 \tab 202 \tab 5 \tab 1599 \cr
   Total \tab 0 \tab 1117 \tab 1822 \tab 66 \tab 429 \tab 25 \tab 3459 \cr
}
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
