% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{milqual}
\alias{milqual}
\title{Quality of persons in military}
\source{
General Social Survey https://gss.norc.org
}
\description{
milqual
}
\details{
Question As you know, this country stopped the military draft in 1972. Since that time we have relied on volunteers. Now I'd like to ask you a few questions about our armed forces.
A. How would you rate the quality of the men and women now serving in the armed forces?  Would you say the quality of personnel is excellent, good, not so good, or poor? (Just your own opinion based on what you've heard or read.)
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} good
\item \code{3} not so good
\item \code{4} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/781/vshow}{GSS Data Explorer page for \code{milqual}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab excellent \tab good \tab not so good \tab poor \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 110 \tab 843 \tab 558 \tab 180 \tab 149 \tab 20 \tab 1860 \cr
   1983 \tab - \tab 134 \tab 877 \tab 355 \tab 108 \tab 121 \tab 4 \tab 1599 \cr
   1984 \tab - \tab 161 \tab 832 \tab 335 \tab 56 \tab 69 \tab 20 \tab 1473 \cr
   Total \tab 0 \tab 405 \tab 2552 \tab 1248 \tab 344 \tab 339 \tab 44 \tab 4932 \cr
}
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
