% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{mindbody}
\alias{mindbody}
\title{Afterlife contrast: spiritual or sensual}
\source{
General Social Survey https://gss.norc.org
}
\description{
mindbody
}
\details{
Question A. People picture life after death in many different ways. We'd like to know how you think of life after death. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of life after death. Look at the first set of contrasting images. If you think that life after death is "A life like the one here on earth only better," you would place yourself at 1. If you think it is "A life without many things which make our present life enjoyable" you would place yourself at 7. If you think it is somewhere between these two you would place yourself at 2, 3, 4, 5, or 6. Where would you place your images of life after death for . . .?
}
\section{Values}{

\itemize{
\item \verb{[1]} spiritual
\item \verb{[7]} sensual life
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/312/vshow}{GSS Data Explorer page for \code{mindbody}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab spiritual \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab sensual life \tab don't know \tab no answer \tab Total \cr
   1986 \tab 246 \tab 478 \tab 136 \tab 126 \tab 216 \tab 42 \tab 37 \tab 143 \tab 19 \tab 27 \tab 1470 \cr
   1987 \tab 369 \tab 554 \tab 132 \tab 147 \tab 221 \tab 86 \tab 58 \tab 179 \tab 22 \tab 51 \tab 1819 \cr
   1988 \tab 282 \tab 432 \tab 135 \tab 115 \tab 222 \tab 57 \tab 53 \tab 131 \tab 15 \tab 39 \tab 1481 \cr
   1989 \tab 749 \tab 247 \tab 82 \tab 62 \tab 159 \tab 48 \tab 42 \tab 99 \tab 15 \tab 34 \tab 1537 \cr
   Total \tab 1646 \tab 1711 \tab 485 \tab 450 \tab 818 \tab 233 \tab 190 \tab 552 \tab 71 \tab 151 \tab 6307 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
