% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{minfood}
\alias{minfood}
\title{Imputed wkly food spending for pov line fam of 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
minfood
}
\details{
Question Now more specifically, thinking about that family of four at the "poverty line," what would you say that family needs to spend each week on food, in this community?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5203/vshow}{GSS Data Explorer page for \code{minfood}}.
}

\seealso{
Other Income: 
\code{\link{minthree}}

Other Income: 
\code{\link{minthree}}
}
\concept{Income}
\keyword{variable}
