% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{minmilop}
\alias{minmilop}
\title{Min opportunities better in military?}
\source{
General Social Survey https://gss.norc.org
}
\description{
minmilop
}
\details{
Question A. Thinking about opportunities and equal treatment for minority groups, like Blacks and Hispanics - Would you say that their treatment and opportunities are better in the military, better in civilian employment, or that there isn't any difference these days?
}
\section{Values}{

\itemize{
\item \verb{[1]} military better
\item \verb{[2]} civilian better
\item \verb{[3]} no difference
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/814/vshow}{GSS Data Explorer page for \code{minmilop}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab military better \tab civilian better \tab no difference \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 706 \tab 85 \tab 558 \tab 100 \tab 24 \tab 1473 \cr
   Total \tab 0 \tab 706 \tab 85 \tab 558 \tab 100 \tab 24 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}
}
\concept{Core}
\keyword{variable}
