% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_220.R
\name{mnildo2}
\alias{mnildo2}
\title{2nd mention what r did when felt mental illness}
\source{
General Social Survey https://gss.norc.org
}
\description{
mnildo2
}
\details{
Question (IF NO TO EVBRKDWN, ASK EVMHP) Have you ever felt you had a mental health problem?
B2. What R. did about the mental illness -- Second mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2083/vshow}{GSS Data Explorer page for \code{mnildo2}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 22 \tab 23 \tab 24 \tab 31 \tab 61 \tab 69 \tab 72 \tab 74 \tab 81 \tab 82 \tab no answer \tab Total \cr
   1996 \tab 2854 \tab 2 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 6 \tab 14 \tab 15 \tab 2904 \cr
   Total \tab 2854 \tab 2 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 6 \tab 14 \tab 15 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{mnildo1}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy3}}
}
\concept{Mental Health}
\keyword{variable}
