% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_220.R
\name{mnilhlp1}
\alias{mnilhlp1}
\title{1st mention what help r sought}
\source{
General Social Survey https://gss.norc.org
}
\description{
mnilhlp1
}
\details{
Question (IF NO TO EVBRKDWN, ASK EVMHP) Have you ever felt you had a mental health problem?
C1. R''s sources of help during mental illness -- First mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2085/vshow}{GSS Data Explorer page for \code{mnilhlp1}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 100 \tab 101 \tab 106 \tab 200 \tab 201 \tab 390 \tab 401 \tab 402 \tab 403 \tab 480 \tab 490 \tab 499 \tab 500 \tab 502 \tab 503 \tab 506 \tab 604 \tab 605 \tab 690 \tab 910 \tab no answer \tab Total \cr
   1996 \tab 2815 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 8 \tab 4 \tab 1 \tab 13 \tab 2 \tab 1 \tab 1 \tab 23 \tab 15 \tab 2904 \cr
   Total \tab 2815 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 8 \tab 4 \tab 1 \tab 13 \tab 2 \tab 1 \tab 1 \tab 23 \tab 15 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy3}}
}
\concept{Mental Health}
\keyword{variable}
