% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_220.R
\name{mnilhlp3}
\alias{mnilhlp3}
\title{3rd mention what help r sought}
\source{
General Social Survey https://gss.norc.org
}
\description{
mnilhlp3
}
\details{
Question (IF NO TO EVBRKDWN, ASK EVMHP) Have you ever felt you had a mental health problem?
C3. R''s sources of help during mental illness -- Third mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2087/vshow}{GSS Data Explorer page for \code{mnilhlp3}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 108 \tab 200 \tab 400 \tab 500 \tab 506 \tab 590 \tab 605 \tab no answer \tab Total \cr
   1996 \tab 2879 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 15 \tab 2904 \cr
   Total \tab 2879 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 15 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mniltime}},
\code{\link{mnilwhy3}}
}
\concept{Mental Health}
\keyword{variable}
