% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_220.R
\name{mniltime}
\alias{mniltime}
\title{When feelings of mental illness happened}
\source{
General Social Survey https://gss.norc.org
}
\description{
mniltime
}
\details{
Question (IF NO TO EVBRKDWN, ASK EVMHP) Have you ever felt you had a mental health problem?
D. The timing of R''s mental illness.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2088/vshow}{GSS Data Explorer page for \code{mniltime}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab no answer \tab Total \cr
   1996 \tab 2815 \tab 13 \tab 16 \tab 19 \tab 13 \tab 28 \tab 2904 \cr
   Total \tab 2815 \tab 13 \tab 16 \tab 19 \tab 13 \tab 28 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mnilwhy3}}
}
\concept{Mental Health}
\keyword{variable}
