% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{mnthsusa}
\alias{mnthsusa}
\title{Lived in u.s., if less than one year 		months:}
\source{
General Social Survey https://gss.norc.org
}
\description{
mnthsusa
}
\details{
Question How long have you lived in the United States? If you have lived in the United States for more than one period during your life, what is the total number of years you have lived in the United States? - Months
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 month
\item \verb{[11]} 11 months
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7298/vshow}{GSS Data Explorer page for \code{mnthsusa}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 1 month \tab 10 \tab 8 \tab 9 \tab 6 \tab don't know \tab no answer \tab skipped on web \tab 11 months \tab 2 \tab 3 \tab 5 \tab 7 \tab Total \cr
   2018 \tab 2344 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3746 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 9 \tab 2 \tab 271 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3520 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 9610 \tab 3 \tab 2 \tab 3 \tab 4 \tab 2 \tab 11 \tab 8 \tab 275 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9924 \cr
}
}

\keyword{variable}
