% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{mntlas1}
\alias{mntlas1}
\title{1st mention, what mental illness is, asserted}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlas1
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
\enumerate{
\item R's asserted references for what mental illness is -- First mention.
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2040/vshow}{GSS Data Explorer page for \code{mntlas1}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 2 \tab 22 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1996 \tab 2196 \tab 84 \tab 2 \tab 106 \tab 3 \tab 58 \tab 12 \tab 16 \tab 18 \tab 36 \tab 1 \tab 37 \tab 48 \tab 104 \tab 4 \tab 4 \tab 78 \tab 53 \tab 44 \tab 2904 \cr
   Total \tab 2196 \tab 84 \tab 2 \tab 106 \tab 3 \tab 58 \tab 12 \tab 16 \tab 18 \tab 36 \tab 1 \tab 37 \tab 48 \tab 104 \tab 4 \tab 4 \tab 78 \tab 53 \tab 44 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}},
\code{\link{spmentl}}
}
\concept{Mental Health}
\keyword{variable}
