% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{mntlas2}
\alias{mntlas2}
\title{2nd mention, what mental illness is, asserted}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlas2
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
2. R's asserted references for what a mental illness is -- Second mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2041/vshow}{GSS Data Explorer page for \code{mntlas2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 13 \tab 14 \tab 16 \tab 18 \tab 2 \tab 20 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1996 \tab 2586 \tab 27 \tab 1 \tab 24 \tab 3 \tab 5 \tab 10 \tab 1 \tab 35 \tab 2 \tab 20 \tab 15 \tab 46 \tab 2 \tab 7 \tab 35 \tab 41 \tab 44 \tab 2904 \cr
   Total \tab 2586 \tab 27 \tab 1 \tab 24 \tab 3 \tab 5 \tab 10 \tab 1 \tab 35 \tab 2 \tab 20 \tab 15 \tab 46 \tab 2 \tab 7 \tab 35 \tab 41 \tab 44 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
