% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_244.R
\name{mntldif}
\alias{mntldif}
\title{How difficult to get mental health care}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntldif
}
\details{
Question How difficult was it to get the treatment that you sought: very difficult, somewhat difficult, or not at all difficult?
}
\section{Values}{

\itemize{
\item \verb{[1]} very difficult
\item \verb{[2]} somewhat diffclt
\item \verb{[3]} not difficult
\item \verb{[8]} know know
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2191/vshow}{GSS Data Explorer page for \code{mntldif}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab very difficult \tab somewhat diffclt \tab not difficult \tab no answer \tab Total \cr
   1998 \tab 2696 \tab 7 \tab 21 \tab 101 \tab 7 \tab 2832 \cr
   Total \tab 2696 \tab 7 \tab 21 \tab 101 \tab 7 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc1}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{mntldeny}}
}
\concept{Mental Health}
\keyword{variable}
