% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{mntlimp2}
\alias{mntlimp2}
\title{2nd mention, what mental illness is, imputed}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlimp2
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
2. R's imputed references for what mental illness is -- Second mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2043/vshow}{GSS Data Explorer page for \code{mntlimp2}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 101 \tab 102 \tab 103 \tab 104 \tab 105 \tab 106 \tab 107 \tab 108 \tab 109 \tab 110 \tab 111 \tab 112 \tab 113 \tab no answer \tab Total \cr
   1996 \tab 2598 \tab 17 \tab 38 \tab 18 \tab 15 \tab 45 \tab 7 \tab 37 \tab 24 \tab 24 \tab 9 \tab 4 \tab 1 \tab 12 \tab 55 \tab 2904 \cr
   Total \tab 2598 \tab 17 \tab 38 \tab 18 \tab 15 \tab 45 \tab 7 \tab 37 \tab 24 \tab 24 \tab 9 \tab 4 \tab 1 \tab 12 \tab 55 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
