% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{mntloth}
\alias{mntloth}
\title{What mental illness is, other codes}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntloth
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
\enumerate{
\item Other mentions about the nature of mental illness.
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2048/vshow}{GSS Data Explorer page for \code{mntloth}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 1007 \tab 1011 \tab 1102 \tab 1106 \tab 1107 \tab 1108 \tab 1110 \tab 1201 \tab 1204 \tab 1205 \tab 1206 \tab 1207 \tab 1209 \tab 1210 \tab 1211 \tab 1214 \tab no answer \tab Total \cr
   1996 \tab 2686 \tab 2 \tab 1 \tab 4 \tab 1 \tab 17 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 14 \tab 51 \tab 4 \tab 53 \tab 42 \tab 18 \tab 2904 \cr
   Total \tab 2686 \tab 2 \tab 1 \tab 4 \tab 1 \tab 17 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 14 \tab 51 \tab 4 \tab 53 \tab 42 \tab 18 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
