% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{mntlrel}
\alias{mntlrel}
\title{Effect of mental health problem on relation with}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlrel
}
\details{
Question I want you to think about the person with a mental health problem with whom you have had the most contact. As a result of the mental health problem, would you say that your relationship with this person...
}
\section{Values}{

\itemize{
\item \verb{[1]} became stronger
\item \verb{[2]} was unchanged
\item \verb{[3]} became worse
\item \verb{[4]} ended as a result of the problem
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2973/vshow}{GSS Data Explorer page for \code{mntlrel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab became stronger \tab was unchanged \tab became worse \tab ended as a result of the problem \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1859 \tab 218 \tab 491 \tab 71 \tab 58 \tab 51 \tab 17 \tab 2765 \cr
   Total \tab 1859 \tab 218 \tab 491 \tab 71 \tab 58 \tab 51 \tab 17 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
