% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{mntlsym1}
\alias{mntlsym1}
\title{1st mention, what mental illness is, symptoms}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlsym1
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
\enumerate{
\item Symptoms or manifestations of mental illness -- First mention.
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2045/vshow}{GSS Data Explorer page for \code{mntlsym1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 202 \tab 203 \tab 204 \tab 205 \tab 206 \tab 207 \tab 208 \tab 210 \tab 211 \tab 212 \tab 215 \tab 303 \tab 304 \tab 306 \tab 308 \tab 310 \tab 402 \tab 405 \tab 407 \tab 409 \tab 411 \tab 501 \tab 502 \tab 503 \tab 512 \tab 601 \tab 603 \tab 611 \tab 701 \tab 702 \tab 704 \tab 707 \tab 708 \tab 709 \tab 710 \tab 712 \tab 804 \tab 807 \tab 808 \tab 810 \tab 811 \tab 812 \tab 902 \tab 906 \tab 907 \tab 908 \tab 909 \tab 921 \tab 922 \tab 923 \tab 924 \tab 925 \tab no answer \tab Total \cr
   1996 \tab 2196 \tab 32 \tab 5 \tab 2 \tab 25 \tab 1 \tab 40 \tab 35 \tab 34 \tab 2 \tab 113 \tab 1 \tab 5 \tab 9 \tab 9 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 13 \tab 12 \tab 57 \tab 15 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 57 \tab 19 \tab 4 \tab 15 \tab 1 \tab 10 \tab 3 \tab 12 \tab 35 \tab 4 \tab 2 \tab 2 \tab 86 \tab 2904 \cr
   Total \tab 2196 \tab 32 \tab 5 \tab 2 \tab 25 \tab 1 \tab 40 \tab 35 \tab 34 \tab 2 \tab 113 \tab 1 \tab 5 \tab 9 \tab 9 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 13 \tab 12 \tab 57 \tab 15 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 57 \tab 19 \tab 4 \tab 15 \tab 1 \tab 10 \tab 3 \tab 12 \tab 35 \tab 4 \tab 2 \tab 2 \tab 86 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
