% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{mntlsym2}
\alias{mntlsym2}
\title{2nd mention, what mental illness is, symptoms}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlsym2
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
2. Symptoms or manifestations of mental illness -- Second mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2046/vshow}{GSS Data Explorer page for \code{mntlsym2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 201 \tab 202 \tab 203 \tab 204 \tab 205 \tab 207 \tab 208 \tab 210 \tab 211 \tab 212 \tab 302 \tab 303 \tab 304 \tab 306 \tab 307 \tab 308 \tab 310 \tab 404 \tab 405 \tab 406 \tab 407 \tab 408 \tab 409 \tab 411 \tab 412 \tab 501 \tab 503 \tab 506 \tab 507 \tab 508 \tab 510 \tab 512 \tab 601 \tab 602 \tab 609 \tab 610 \tab 611 \tab 612 \tab 701 \tab 702 \tab 704 \tab 707 \tab 708 \tab 709 \tab 710 \tab 711 \tab 712 \tab 801 \tab 802 \tab 803 \tab 805 \tab 807 \tab 808 \tab 810 \tab 811 \tab 812 \tab 902 \tab 904 \tab 905 \tab 906 \tab 908 \tab 909 \tab 911 \tab 921 \tab 922 \tab 923 \tab 925 \tab no answer \tab Total \cr
   1996 \tab 2391 \tab 3 \tab 11 \tab 4 \tab 2 \tab 10 \tab 31 \tab 29 \tab 27 \tab 4 \tab 28 \tab 1 \tab 14 \tab 16 \tab 8 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab 3 \tab 2 \tab 6 \tab 8 \tab 4 \tab 13 \tab 16 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 8 \tab 2 \tab 33 \tab 12 \tab 10 \tab 3 \tab 6 \tab 12 \tab 11 \tab 1 \tab 1 \tab 5 \tab 20 \tab 5 \tab 1 \tab 86 \tab 2904 \cr
   Total \tab 2391 \tab 3 \tab 11 \tab 4 \tab 2 \tab 10 \tab 31 \tab 29 \tab 27 \tab 4 \tab 28 \tab 1 \tab 14 \tab 16 \tab 8 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab 3 \tab 2 \tab 6 \tab 8 \tab 4 \tab 13 \tab 16 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 8 \tab 2 \tab 33 \tab 12 \tab 10 \tab 3 \tab 6 \tab 12 \tab 11 \tab 1 \tab 1 \tab 5 \tab 20 \tab 5 \tab 1 \tab 86 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
