% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{mobiledata}
\alias{mobiledata}
\title{Household has access to internet using cellular data plan}
\source{
General Social Survey https://gss.norc.org
}
\description{
mobiledata
}
\details{
Question Do you or any member of this household have access to the Internet using a cellular data plan for a smartphone or other mobile device?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8196/vshow}{GSS Data Explorer page for \code{mobiledata}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 457 \tab 150 \tab 4 \tab 2697 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 457 \tab 150 \tab 4 \tab 2697 \tab 1 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{broadband}},
\code{\link{spnorelgsp}}
}
\concept{Core}
\keyword{variable}
