% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{modact}
\alias{modact}
\title{R limited in moderate activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
modact
}
\details{
Question The following questions are about activities you might do during a typical day. Does your health now limit you in these activities?  If so, how much?
A. Moderate activities, such as moving a table, pushing a vacuum cleaner, bowling, or playing golf?
}
\section{Values}{

\itemize{
\item \code{1} yes, limited a lot
\item \code{2} yes, limited a little
\item \code{3} no, not limited at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2677/vshow}{GSS Data Explorer page for \code{modact}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, limited a lot \tab yes, limited a little \tab no, not limited at all \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 136 \tab 159 \tab 1103 \tab 21 \tab 2817 \cr
   Total \tab 1398 \tab 136 \tab 159 \tab 1103 \tab 21 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{crelesse}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{limitedp}},
\code{\link{stairs}}
}
\concept{Health Status}
\keyword{variable}
