% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{mode}
\alias{mode}
\title{Interview done in-person or over the phone}
\source{
General Social Survey https://gss.norc.org
}
\description{
mode
}
\details{
Question Interview done in-person or over the phone
}
\section{Values}{

\itemize{
\item \verb{[1]} in-person
\item \verb{[2]} by phone
\item \verb{[3]} multimode
\item \verb{[4]} web
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5566/vshow}{GSS Data Explorer page for \code{mode}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab in-person \tab by phone \tab don't know \tab no answer \tab multimode \tab web \tab Total \cr
   2004 \tab - \tab 2219 \tab 593 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4209 \tab 301 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1778 \tab 245 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1824 \tab 211 \tab 1 \tab 8 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1712 \tab 258 \tab - \tab 4 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2199 \tab 333 \tab - \tab 6 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2665 \tab 195 \tab - \tab 7 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2094 \tab 245 \tab - \tab 9 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 293 \tab - \tab - \tab 218 \tab 3521 \tab 4032 \cr
   2022 \tab - \tab 1562 \tab 201 \tab - \tab - \tab 148 \tab 1633 \tab 3544 \cr
   2024 \tab - \tab 1174 \tab 254 \tab - \tab - \tab 119 \tab 1762 \tab 3309 \cr
   Total \tab 0 \tab 21436 \tab 3129 \tab 1 \tab 34 \tab 485 \tab 6916 \tab 32001 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{devtype}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{huclean}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{intid}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{intyrs_0418}},
\code{\link{lngthinv}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse1}},
\code{\link{whoelse2}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}

Other Interview: 
\code{\link{comprend}},
\code{\link{consent}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{id}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{inthisp}},
\code{\link{intid}},
\code{\link{intrace1}},
\code{\link{intrace2}},
\code{\link{intrace3}},
\code{\link{intrwght}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{intyrs_0418}},
\code{\link{lngthinv}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{sciintro}},
\code{\link{scimode}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{year}()}
}
\concept{Interview}
\concept{Paradata}
\keyword{variable}
