% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{modelobby}
\alias{modelobby}
\title{Has r contacted a politician in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
modelobby
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the last 12 months offline only, online only, or both offline and online.... Contacted, or attempted to contact, a politician to express your views.
}
\section{Values}{

\itemize{
\item \verb{[1]} only offline
\item \verb{[2]} only online
\item \verb{[3]} both offline and online
\item \verb{[4]} neither offline nor online
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8153/vshow}{GSS Data Explorer page for \code{modelobby}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab only offline \tab only online \tab both offline and online \tab neither offline nor online \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 54 \tab 178 \tab 90 \tab 1189 \tab 104 \tab 1671 \tab 10 \tab 13 \tab 3309 \cr
   Total \tab 0 \tab 54 \tab 178 \tab 90 \tab 1189 \tab 104 \tab 1671 \tab 10 \tab 13 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{polnewsfrom}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
