% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{modeorgprot}
\alias{modeorgprot}
\title{Has r organized a protest in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
modeorgprot
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the last 12 months offline only, online only, or both offline and online… Organized or helped to organize a demonstration or political protest.
}
\section{Values}{

\itemize{
\item \verb{[1]} only offline
\item \verb{[2]} only online
\item \verb{[3]} both offline and online
\item \verb{[4]} neither offline nor online
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8155/vshow}{GSS Data Explorer page for \code{modeorgprot}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab only offline \tab only online \tab both offline and online \tab neither offline nor online \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 20 \tab 18 \tab 32 \tab 1443 \tab 97 \tab 1671 \tab 11 \tab 17 \tab 3309 \cr
   Total \tab 0 \tab 20 \tab 18 \tab 32 \tab 1443 \tab 97 \tab 1671 \tab 11 \tab 17 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{polnewsfrom}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
