% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{modesequence}
\alias{modesequence}
\title{Initial mode of contact for respondent}
\source{
General Social Survey https://gss.norc.org
}
\description{
modesequence
}
\details{
Question SEQUENCE OF MODE ASSIGNMENT FOR RESPONDENT
}
\section{Values}{

\itemize{
\item \code{1} in-person first
\item \code{2} web-first
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7970/vshow}{GSS Data Explorer page for \code{modesequence}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab in-person first \tab web-first \tab Total \cr
   2022 \tab - \tab 1540 \tab 2004 \tab 3544 \cr
   Total \tab 0 \tab 1540 \tab 2004 \tab 3544 \cr
}
}

\keyword{variable}
