% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{monday}
\alias{monday}
\title{Attended religious services last monday}
\source{
General Social Survey https://gss.norc.org
}
\description{
monday
}
\details{
Question On what day or days do you attend religious services during the last seven days? (PROBE, ASK UNTIL "NO" IS GIVEN: "Did you attend religious services on any other days during the last seven days?")
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5312/vshow}{GSS Data Explorer page for \code{monday}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2608 \tab 15 \tab 195 \tab 86 \tab 2904 \cr
   1998 \tab 2329 \tab 24 \tab 379 \tab 100 \tab 2832 \cr
   Total \tab 4937 \tab 39 \tab 574 \tab 186 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attrelig}},
\code{\link{eatout}},
\code{\link{godoc}},
\code{\link{numdays}},
\code{\link{seefilm}},
\code{\link{sunday}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
