% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{monitred}
\alias{monitred}
\title{Computers monitor my job performance}
\source{
General Social Survey https://gss.norc.org
}
\description{
monitred
}
\details{
Question How true would you say each of the following statements is about your job?  Would you say very true, true, somewhat true, or not true at all:
E. Computers monitor my job performance.
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} true
\item \verb{[3]} somewhat true
\item \verb{[4]} not true at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1312/vshow}{GSS Data Explorer page for \code{monitred}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very true \tab true \tab somewhat true \tab not true at all \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 605 \tab 35 \tab 59 \tab 66 \tab 736 \tab 4 \tab 12 \tab 1517 \cr
   Total \tab 605 \tab 35 \tab 59 \tab 66 \tab 736 \tab 4 \tab 12 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{totsup}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
