% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_202.R
\name{monwork}
\alias{monwork}
\title{Done volunteer work in work-related org}
\source{
General Social Survey https://gss.norc.org
}
\description{
monwork
}
\details{
Question In which, if any, of the areas listed on this card have you done some volunteer work in the past month? (RECORD above UNDER MONHLTH-MONOTH. CIRCLE "YES" OR "NO" FOR EACH AREA)
I. Work-related organizations
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1908/vshow}{GSS Data Explorer page for \code{monwork}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2733 \tab 86 \tab 72 \tab 13 \tab 2904 \cr
   Total \tab 2733 \tab 86 \tab 72 \tab 13 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{monart}},
\code{\link{monenvir}},
\code{\link{monfound}},
\code{\link{monhuman}},
\code{\link{monpol}},
\code{\link{monpub}},
\code{\link{monrec}},
\code{\link{monrelig}},
\code{\link{monyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
