% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_202.R
\name{monyouth}
\alias{monyouth}
\title{H done volunteer work in youth development}
\source{
General Social Survey https://gss.norc.org
}
\description{
monyouth
}
\details{
Question In which, if any, of the areas listed on this card have you done some volunteer work in the past month? (RECORD above UNDER MONHLTH-MONOTH. CIRCLE "YES" OR "NO" FOR EACH AREA)
K. Youth development
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1910/vshow}{GSS Data Explorer page for \code{monyouth}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2690 \tab 107 \tab 95 \tab 12 \tab 2904 \cr
   Total \tab 2690 \tab 107 \tab 95 \tab 12 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{monart}},
\code{\link{monenvir}},
\code{\link{monfound}},
\code{\link{monhuman}},
\code{\link{monpol}},
\code{\link{monpub}},
\code{\link{monrec}},
\code{\link{monrelig}},
\code{\link{monwork}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
