% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{moonlite}
\alias{moonlite}
\title{Does r have a second job}
\source{
General Social Survey https://gss.norc.org
}
\description{
moonlite
}
\details{
Question And aside from your main job, do you have any other paid jobs, like a second job or other paid work?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, regularly
\item \verb{[2]} yes, sometimes
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4564/vshow}{GSS Data Explorer page for \code{moonlite}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab yes, regularly \tab yes, sometimes \tab no \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 71 \tab 158 \tab 644 \tab 664 \tab - \tab 1537 \cr
   Total \tab 0 \tab 71 \tab 158 \tab 644 \tab 664 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{totalnum}}
}
\concept{Work Orientation}
\keyword{variable}
