% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{moregood}
\alias{moregood}
\title{I expect more good things to happen to me than bad}
\source{
General Social Survey https://gss.norc.org
}
\description{
moregood
}
\details{
Question Indicate your agreement with each of the following statements by selecting the number that comes closest to your answer:
I. Overall, I expect more good things to happen to me than bad.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3044/vshow}{GSS Data Explorer page for \code{moregood}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2004 \tab 432 \tab 819 \tab 1308 \tab 205 \tab 39 \tab 6 \tab 3 \tab 2812 \cr
   Total \tab 432 \tab 819 \tab 1308 \tab 205 \tab 39 \tab 6 \tab 3 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{owndoing}},
\code{\link{sadblue}},
\code{\link{selfish}}
}
\concept{Genes and the Environment}
\concept{Work Organizations}
\keyword{variable}
