% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{morempg}
\alias{morempg}
\title{Cars use less gasoline}
\source{
General Social Survey https://gss.norc.org
}
\description{
morempg
}
\details{
Question How much do you favor or oppose requiring car makers to make cars and trucks that use less gasoline? Do you strongly favor, favor, neither favor nor oppose, oppose, or strongly oppose such a requirement?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3487/vshow}{GSS Data Explorer page for \code{morempg}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1569 \tab 260 \tab 138 \tab 48 \tab 13 \tab 10 \tab 2 \tab 4 \tab 2044 \cr
   Total \tab 1569 \tab 260 \tab 138 \tab 48 \tab 13 \tab 10 \tab 2 \tab 4 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
