% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{morenuke}
\alias{morenuke}
\title{Possibility of increasing arms race}
\source{
General Social Survey https://gss.norc.org
}
\description{
morenuke
}
\details{
Question I'm going to read you some possible military situations the U.S. might face in the next ten years. (HAND CARD BB)
Some people feel these situations are certain to happen (think of these as point 7 on the scale), others think these situations won't happen at all (think of these as point 1 on the scale). And of course some people have opinions somewhere in between. For each of these possible military situations, please give me your best guess as to how likely it is to happen. For example, an all-out atomic war. Where would you put the likelihood of an all-out atomic war during the next ten years? (Would you say 7--it is certain to happen, 1--it won't happen at all, or something in between?)
(CIRCLE ONE CODE FOR "A" AND REPEAT FOR B-F)
C. Peace but increasing arms build by the U.S. and Russia?
}
\section{Values}{

\itemize{
\item \verb{[1]} wont happen
\item \verb{[7]} certain to happen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/824/vshow}{GSS Data Explorer page for \code{morenuke}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab wont happen \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab certain to happen \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 91 \tab 83 \tab 112 \tab 319 \tab 296 \tab 227 \tab 278 \tab 43 \tab 24 \tab 1473 \cr
   Total \tab 0 \tab 91 \tab 83 \tab 112 \tab 319 \tab 296 \tab 227 \tab 278 \tab 43 \tab 24 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{defwrkev}},
\code{\link{guerilla}},
\code{\link{jobtrain}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()},
\code{\link{vetaid}}
}
\concept{Core}
\keyword{variable}
