% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{morextrm}
\alias{morextrm}
\title{Everyone is like x but x's case is just a bit more extreme and ongoing}
\source{
General Social Survey https://gss.norc.org
}
\description{
morextrm
}
\details{
Question Let’s continue talking about (Name) for a minute. Tell me how much you agree or disagree with the following statement. Basically, we are all a little like (Name). In (Name)’s case it is just a bit more extreme and ongoing.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8044/vshow}{GSS Data Explorer page for \code{morextrm}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 170 \tab 792 \tab 454 \tab 125 \tab 33 \tab 1693 \tab 15 \tab 27 \tab 3309 \cr
   Total \tab 0 \tab 170 \tab 792 \tab 454 \tab 125 \tab 33 \tab 1693 \tab 15 \tab 27 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{bullied}},
\code{\link{imprvown1}},
\code{\link{mustdoc1}},
\code{\link{musthosp1}},
\code{\link{mustmed1}},
\code{\link{smedia}}
}
\concept{Children}
\concept{Mental Health}
\keyword{variable}
