% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{mostact}
\alias{mostact}
\title{Which group is r most active in}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostact
}
\details{
Question Which one of these groups or organizations (READ GROUP FROM MEMFRAT THAT R REPLIED "YES" TO) are you most active in?  CODE ONLY ONE.
}
\section{Values}{

\itemize{
\item \verb{[1]} fraternal groups
\item \verb{[2]} service clubs
\item \verb{[3]} veterans groups
\item \verb{[4]} political clubs
\item \verb{[5]} labor unions
\item \verb{[6]} sports groups
\item \verb{[7]} youth groups
\item \verb{[8]} school service groups
\item \verb{[9]} hobby or garden clubs
\item \verb{[10]} school fraternities or sororities
\item \verb{[11]} nationality groups
\item \verb{[12]} farm organizations
\item \verb{[13]} literary art discussion or study grps
\item \verb{[14]} professional or academic societies
\item \verb{[15]} church-affiliated groups
\item \verb{[16]} other groups
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1013/vshow}{GSS Data Explorer page for \code{mostact}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab fraternal groups \tab service clubs \tab veterans groups \tab political clubs \tab labor unions \tab sports groups \tab youth groups \tab school service groups \tab hobby or garden clubs \tab school fraternities or sororities \tab nationality groups \tab farm organizations \tab literary art discussion or study grps \tab professional or academic societies \tab church-affiliated groups \tab other groups \tab no answer \tab Total \cr
   1987 \tab 1083 \tab 40 \tab 32 \tab 21 \tab 9 \tab 25 \tab 107 \tab 30 \tab 48 \tab 31 \tab 9 \tab 7 \tab 12 \tab 18 \tab 46 \tab 224 \tab 38 \tab 39 \tab 1819 \cr
   Total \tab 1083 \tab 40 \tab 32 \tab 21 \tab 9 \tab 25 \tab 107 \tab 30 \tab 48 \tab 31 \tab 9 \tab 7 \tab 12 \tab 18 \tab 46 \tab 224 \tab 38 \tab 39 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{servegrp}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
