% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{mosthrs}
\alias{mosthrs}
\title{Most hrs/week worked in past month}
\source{
General Social Survey https://gss.norc.org
}
\description{
mosthrs
}
\details{
Question In the last month, what is the greatest number of hours you worked in a week, at all paid jobs? Please consider all hours, including any extra hours, overtime, work you did at home for your job, and time you spent on work that may not have been directly billable or compensated.
HOURS:
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6293/vshow}{GSS Data Explorer page for \code{mosthrs}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 100 \tab 12 \tab 120 \tab 145 \tab 15 \tab 16 \tab 18 \tab 19 \tab 2 \tab 20 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 60 \tab 62 \tab 63 \tab 64 \tab 65 \tab 68 \tab 7 \tab 70 \tab 72 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 8 \tab 80 \tab 82 \tab 84 \tab 86 \tab 88 \tab 90 \tab 92 \tab 98 \tab 99 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2273 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 1 \tab 9 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 2 \tab 3 \tab 1 \tab 26 \tab 1 \tab 8 \tab 1 \tab 1 \tab 14 \tab 4 \tab 3 \tab 2 \tab 7 \tab 104 \tab 7 \tab 20 \tab 5 \tab 8 \tab 30 \tab 6 \tab 1 \tab 19 \tab 1 \tab 1 \tab 64 \tab 9 \tab 2 \tab 2 \tab 14 \tab 5 \tab 1 \tab 3 \tab 1 \tab 2 \tab 47 \tab 5 \tab 2 \tab 4 \tab 16 \tab 4 \tab 1 \tab 30 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 10 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 20 \tab 2867 \cr
   Total \tab 2273 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 1 \tab 9 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 2 \tab 3 \tab 1 \tab 26 \tab 1 \tab 8 \tab 1 \tab 1 \tab 14 \tab 4 \tab 3 \tab 2 \tab 7 \tab 104 \tab 7 \tab 20 \tab 5 \tab 8 \tab 30 \tab 6 \tab 1 \tab 19 \tab 1 \tab 1 \tab 64 \tab 9 \tab 2 \tab 2 \tab 14 \tab 5 \tab 1 \tab 3 \tab 1 \tab 2 \tab 47 \tab 5 \tab 2 \tab 4 \tab 16 \tab 4 \tab 1 \tab 30 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 10 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 20 \tab 2867 \cr
}
}

\seealso{
Other Flexible Work: 
\code{\link{paidhow}},
\code{\link{usualhrs}}
}
\concept{Flexible Work}
\keyword{variable}
