% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_153.R
\name{mostimp2}
\alias{mostimp2}
\title{2nd most imp event in lst 60 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostimp2
}
\details{
Question The next questions concern how people think about the past. There have been a lot of national and world events  and changes over the past 60 years--say, from about 1930 right up until today. Would you mention one or two such events or changes that seem to you to have been especially important? There aren't any right or wrong answers to the question--just whatever national or world events or changes over the past 60 years that come to mind as important to you.
}
\section{Values}{

\itemize{
\item \code{990} other
\item \code{995} personal
\item \code{996} none, no event
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1447/vshow}{GSS Data Explorer page for \code{mostimp2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 141 \tab 143 \tab 144 \tab 150 \tab 18 \tab 19 \tab 2 \tab 21 \tab 22 \tab 23 \tab 25 \tab 26 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 55 \tab 6 \tab 60 \tab 61 \tab 62 \tab 64 \tab 66 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 75 \tab 76 \tab 77 \tab 79 \tab other \tab personal \tab don't know \tab no answer \tab Total \cr
   1993 \tab 226 \tab 17 \tab 3 \tab 2 \tab 11 \tab 7 \tab 28 \tab 24 \tab 45 \tab 23 \tab 3 \tab 4 \tab 1 \tab 20 \tab 13 \tab 36 \tab 4 \tab 32 \tab 95 \tab 3 \tab 4 \tab 18 \tab 6 \tab 2 \tab 1 \tab 44 \tab 16 \tab 4 \tab 62 \tab 15 \tab 8 \tab 1 \tab 4 \tab 19 \tab 190 \tab 48 \tab 6 \tab 2 \tab 22 \tab 106 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 11 \tab 1 \tab 1 \tab 77 \tab 5 \tab 61 \tab 7 \tab 1 \tab 9 \tab 22 \tab 84 \tab 28 \tab 94 \tab 21 \tab 1606 \cr
   Total \tab 226 \tab 17 \tab 3 \tab 2 \tab 11 \tab 7 \tab 28 \tab 24 \tab 45 \tab 23 \tab 3 \tab 4 \tab 1 \tab 20 \tab 13 \tab 36 \tab 4 \tab 32 \tab 95 \tab 3 \tab 4 \tab 18 \tab 6 \tab 2 \tab 1 \tab 44 \tab 16 \tab 4 \tab 62 \tab 15 \tab 8 \tab 1 \tab 4 \tab 19 \tab 190 \tab 48 \tab 6 \tab 2 \tab 22 \tab 106 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 11 \tab 1 \tab 1 \tab 77 \tab 5 \tab 61 \tab 7 \tab 1 \tab 9 \tab 22 \tab 84 \tab 28 \tab 94 \tab 21 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{impcultr}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{mostimp1}},
\code{\link{mostimp3}}
}
\concept{Culture}
\keyword{variable}
