% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_153.R
\name{mostimp3}
\alias{mostimp3}
\title{3rd imp event in lst 60 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostimp3
}
\details{
Question The next questions concern how people think about the past. There have been a lot of national and world events  and changes over the past 60 years--say, from about 1930 right up until today. Would you mention one or two such events or changes that seem to you to have been especially important? There aren't any right or wrong answers to the question--just whatever national or world events or changes over the past 60 years that come to mind as important to you.
}
\section{Values}{

\itemize{
\item \verb{[990]} other
\item \verb{[995]} personal
\item \verb{[996]} none, no event
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1448/vshow}{GSS Data Explorer page for \code{mostimp3}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 150 \tab 18 \tab 19 \tab 2 \tab 21 \tab 26 \tab 29 \tab 3 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 48 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 6 \tab 61 \tab 62 \tab 64 \tab 65 \tab 66 \tab 69 \tab 71 \tab 72 \tab 73 \tab 77 \tab 79 \tab other \tab personal \tab don't know \tab no answer \tab Total \cr
   1993 \tab 1186 \tab 3 \tab 2 \tab 7 \tab 8 \tab 19 \tab 3 \tab 9 \tab 2 \tab 15 \tab 7 \tab 26 \tab 3 \tab 2 \tab 5 \tab 2 \tab 3 \tab 11 \tab 16 \tab 2 \tab 20 \tab 8 \tab 6 \tab 3 \tab 3 \tab 1 \tab 41 \tab 13 \tab 1 \tab 2 \tab 1 \tab 22 \tab 31 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 17 \tab 3 \tab 23 \tab 2 \tab 2 \tab 27 \tab 9 \tab 8 \tab 21 \tab 1606 \cr
   Total \tab 1186 \tab 3 \tab 2 \tab 7 \tab 8 \tab 19 \tab 3 \tab 9 \tab 2 \tab 15 \tab 7 \tab 26 \tab 3 \tab 2 \tab 5 \tab 2 \tab 3 \tab 11 \tab 16 \tab 2 \tab 20 \tab 8 \tab 6 \tab 3 \tab 3 \tab 1 \tab 41 \tab 13 \tab 1 \tab 2 \tab 1 \tab 22 \tab 31 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 17 \tab 3 \tab 23 \tab 2 \tab 2 \tab 27 \tab 9 \tab 8 \tab 21 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{impcultr}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{mostimp1}},
\code{\link{mostimp2}}
}
\concept{Culture}
\keyword{variable}
