% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_38.R
\name{mother}
\alias{mother}
\title{Image of god as mother}
\source{
General Social Survey https://gss.norc.org
}
\description{
mother
}
\details{
Question When you think about God, how likely are each of these images to come to your mind?  Would you say extremely likely, somewhat likely, not too likely, or not likely at all? (HAND CARD AD. READ EACH ITEM.) (CIRCLE ONE NUMBER FOR EACH WORD)
E. Mother
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not too likely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/326/vshow}{GSS Data Explorer page for \code{mother}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab extremely likely \tab somewhat likely \tab not too likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab - \tab 389 \tab 368 \tab 365 \tab 406 \tab 52 \tab 19 \tab 1599 \cr
   1984 \tab 989 \tab 105 \tab 113 \tab 94 \tab 147 \tab 14 \tab 11 \tab 1473 \cr
   Total \tab 989 \tab 494 \tab 481 \tab 459 \tab 553 \tab 66 \tab 30 \tab 3072 \cr
}
}

\seealso{
Other Core: 
\code{\link{creator}},
\code{\link{father}},
\code{\link{friend}},
\code{\link{judge}},
\code{\link{lover}},
\code{\link{master}},
\code{\link{neargod}},
\code{\link{redeemer}},
\code{\link{spouse}}
}
\concept{Core}
\keyword{variable}
