% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{movetown}
\alias{movetown}
\title{Willing to move to another town or city}
\source{
General Social Survey https://gss.norc.org
}
\description{
movetown
}
\details{
Question If you could improve your work or living conditions, how willing or unwilling would you be to…
B. Move to another town or city within this state
}
\section{Values}{

\itemize{
\item \code{1} very willing
\item \code{2} fairly willing
\item \code{3} neither willing nor unwilling
\item \code{4} fairly unwilling
\item \code{5} very unwilling
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4822/vshow}{GSS Data Explorer page for \code{movetown}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very willing \tab fairly willing \tab neither willing nor unwilling \tab fairly unwilling \tab very unwilling \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 313 \tab 448 \tab 159 \tab 174 \tab 210 \tab 39 \tab 24 \tab 2904 \cr
   Total \tab 1537 \tab 313 \tab 448 \tab 159 \tab 174 \tab 210 \tab 39 \tab 24 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{moveusa}},
\code{\link{onenatn}}
}
\concept{ISSP National Identity}
\keyword{variable}
