% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{moveusa}
\alias{moveusa}
\title{Willing to move outside america}
\source{
General Social Survey https://gss.norc.org
}
\description{
moveusa
}
\details{
Question If you could improve your work or living conditions, how willing or unwilling would you be to…
D. Move outside America
}
\section{Values}{

\itemize{
\item \code{1} very willing
\item \code{2} fairly willing
\item \code{3} neither willing nor unwilling
\item \code{4} fairly unwilling
\item \code{5} very unwilling
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4824/vshow}{GSS Data Explorer page for \code{moveusa}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{onenatn}}

Other Residence: 
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}}
}
\concept{ISSP National Identity}
\concept{Residence}
\keyword{variable}
