% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{mp3}
\alias{mp3}
\title{How familiar r is with term mp3}
\source{
General Social Survey https://gss.norc.org
}
\description{
mp3
}
\details{
Question F. MP3
}
\section{Values}{

\itemize{
\item \verb{[1]} very familiar
\item \verb{[2]} somewhat familiar
\item \verb{[3]} not familiar
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2635/vshow}{GSS Data Explorer page for \code{mp3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very familiar \tab somewhat familiar \tab not familiar \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2142 \tab 163 \tab 141 \tab 271 \tab 1 \tab 47 \tab 2765 \cr
   Total \tab 2142 \tab 163 \tab 141 \tab 271 \tab 1 \tab 47 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{browser3}},
\code{\link{ezines}},
\code{\link{newsgrps}},
\code{\link{prefsets}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}}
}
\concept{Information Society}
\keyword{variable}
