% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_362.R
\name{mtfarm1}
\alias{mtfarm1}
\title{Farm organization1 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtfarm1
}
\details{
Question L. Does the farm organization meet in this area?
\enumerate{
\item First farm organization
}
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3269/vshow}{GSS Data Explorer page for \code{mtfarm1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2763 \tab 25 \tab 18 \tab 6 \tab 2812 \cr
   Total \tab 2763 \tab 25 \tab 18 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
