% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_362.R
\name{mtgreek2}
\alias{mtgreek2}
\title{School frarternity2 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtgreek2
}
\details{
Question J. Does the school fraternity or sorority meet in this area?
2. Second school fraternity or sorority
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3264/vshow}{GSS Data Explorer page for \code{mtgreek2}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no \tab no answer \tab Total \cr
   2004 \tab 2804 \tab 2 \tab 6 \tab 2812 \cr
   Total \tab 2804 \tab 2 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtgreek3}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
