% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_362.R
\name{mtgreek3}
\alias{mtgreek3}
\title{School frarternity3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtgreek3
}
\details{
Question J. Does the school fraternity or sorority meet in this area?
3. Third school fraternity or sorority
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3265/vshow}{GSS Data Explorer page for \code{mtgreek3}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab no answer \tab Total \cr
   2004 \tab 2806 \tab 6 \tab 2812 \cr
   Total \tab 2806 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtgreek2}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
