% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_365.R
\name{mtinfrm3}
\alias{mtinfrm3}
\title{Informal group3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtinfrm3
}
\details{
Question Q. Do any of the other informal groups meet in this area?
3. Third other group?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3295/vshow}{GSS Data Explorer page for \code{mtinfrm3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2767 \tab 26 \tab 5 \tab 1 \tab 13 \tab 2812 \cr
   Total \tab 2767 \tab 26 \tab 5 \tab 1 \tab 13 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet1}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
