% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_363.R
\name{mtlit3}
\alias{mtlit3}
\title{Military group3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtlit3
}
\details{
Question M. Does the literary, art, discussion, or study group meet in this area?
3. Third literary, art, discussion, or study group
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3274/vshow}{GSS Data Explorer page for \code{mtlit3}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtlit4}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}}

Other Voluntary Associations: 
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtlit4}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
