% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_362.R
\name{mtnat3}
\alias{mtnat3}
\title{Nationality group3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtnat3
}
\details{
Question K. Does the nationality group meet in this area?
3. Third nationality group
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3268/vshow}{GSS Data Explorer page for \code{mtnat3}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab no answer \tab Total \cr
   2004 \tab 2806 \tab 6 \tab 2812 \cr
   Total \tab 2806 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtnat1}},
\code{\link{mtnat2}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
