% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_364.R
\name{mtother2}
\alias{mtother2}
\title{Other group2 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtother2
}
\details{
Question P. Do any of the other groups meet in this area?
2. Second other group
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3289/vshow}{GSS Data Explorer page for \code{mtother2}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtinfrm1}},
\code{\link{mtother1}},
\code{\link{mtother3}},
\code{\link{mtother4}},
\code{\link{mtother5}}

Other Voluntary Associations: 
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtinfrm1}},
\code{\link{mtother1}},
\code{\link{mtother3}},
\code{\link{mtother4}},
\code{\link{mtother5}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
