% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_364.R
\name{mtother4}
\alias{mtother4}
\title{Other group4 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtother4
}
\details{
Question P. Do any of the other groups meet in this area?
4. Fourth other group
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3291/vshow}{GSS Data Explorer page for \code{mtother4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2801 \tab 1 \tab 1 \tab 9 \tab 2812 \cr
   Total \tab 2801 \tab 1 \tab 1 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtinfrm1}},
\code{\link{mtother1}},
\code{\link{mtother2}},
\code{\link{mtother3}},
\code{\link{mtother5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
