% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_363.R
\name{mtprof2}
\alias{mtprof2}
\title{Professional society2 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtprof2
}
\details{
Question N. Does the professional or academic society meet in this area?
2. Second professional or academic society
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3277/vshow}{GSS Data Explorer page for \code{mtprof2}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtlit3}},
\code{\link{mtlit4}},
\code{\link{mtprof1}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
