% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_363.R
\name{mtprof5}
\alias{mtprof5}
\title{Professional society5 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtprof5
}
\details{
Question N. Does the professional or academic society meet in this area?
5. Fifth professional or academic society
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3280/vshow}{GSS Data Explorer page for \code{mtprof5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2797 \tab 3 \tab 1 \tab 11 \tab 2812 \cr
   Total \tab 2797 \tab 3 \tab 1 \tab 11 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtlit3}},
\code{\link{mtlit4}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof6}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
