% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_361.R
\name{mtschl2}
\alias{mtschl2}
\title{School service group2 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtschl2
}
\details{
Question H. Does the school service group meet in this area?
2. Second school service group
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3257/vshow}{GSS Data Explorer page for \code{mtschl2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2785 \tab 16 \tab 2 \tab 9 \tab 2812 \cr
   Total \tab 2785 \tab 16 \tab 2 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtgreek1}},
\code{\link{mthobby1}},
\code{\link{mthobby2}},
\code{\link{mthobby3}},
\code{\link{mtschl1}},
\code{\link{mtschl3}},
\code{\link{mtschl4}},
\code{\link{mtyouth4}},
\code{\link{mtyouth5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
