% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_361.R
\name{mtschl3}
\alias{mtschl3}
\title{School service group3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtschl3
}
\details{
Question H. Does the school service group meet in this area?
3. Third school service group
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3258/vshow}{GSS Data Explorer page for \code{mtschl3}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtgreek1}},
\code{\link{mthobby1}},
\code{\link{mthobby2}},
\code{\link{mthobby3}},
\code{\link{mtschl1}},
\code{\link{mtschl2}},
\code{\link{mtschl4}},
\code{\link{mtyouth4}},
\code{\link{mtyouth5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
