% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_358.R
\name{mtserv7}
\alias{mtserv7}
\title{Service club7 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtserv7
}
\details{
Question B. Does the service club meet in this area?
7. Seventh service club
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3231/vshow}{GSS Data Explorer page for \code{mtserv7}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv8}},
\code{\link{mtvet1}}

Other Voluntary Associations: 
\code{\link{mtfrat4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv8}},
\code{\link{mtvet1}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
