% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_358.R
\name{mtserv8}
\alias{mtserv8}
\title{Service club8 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtserv8
}
\details{
Question B. Does the service club meet in this area?
8. Eight service club
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3232/vshow}{GSS Data Explorer page for \code{mtserv8}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no \tab no answer \tab Total \cr
   2004 \tab 2805 \tab 1 \tab 6 \tab 2812 \cr
   Total \tab 2805 \tab 1 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv7}},
\code{\link{mtvet1}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
