% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_360.R
\name{mtsport1}
\alias{mtsport1}
\title{Sport club1 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtsport1
}
\details{
Question F. Does the sports group meet in this area?
\enumerate{
\item First sports group
}
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3245/vshow}{GSS Data Explorer page for \code{mtsport1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2561 \tab 207 \tab 35 \tab 9 \tab 2812 \cr
   Total \tab 2561 \tab 207 \tab 35 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtsport2}},
\code{\link{mtsport3}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion4}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
