% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_360.R
\name{mtsport2}
\alias{mtsport2}
\title{Sport club2 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtsport2
}
\details{
Question F. Does the sports group meet in this area?
2. Second sports group
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3246/vshow}{GSS Data Explorer page for \code{mtsport2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2740 \tab 59 \tab 4 \tab 9 \tab 2812 \cr
   Total \tab 2740 \tab 59 \tab 4 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtsport1}},
\code{\link{mtsport3}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion4}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
