% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_358.R
\name{mtvet1}
\alias{mtvet1}
\title{Veteran group1 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtvet1
}
\details{
Question C. Does the veterans'' group meet in this area?
\enumerate{
\item First veterans'' group
}
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3233/vshow}{GSS Data Explorer page for \code{mtvet1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2729 \tab 57 \tab 19 \tab 7 \tab 2812 \cr
   Total \tab 2729 \tab 57 \tab 19 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv7}},
\code{\link{mtserv8}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
