% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_359.R
\name{mtvet2}
\alias{mtvet2}
\title{Veteran group2 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtvet2
}
\details{
Question C. Does the veterans'' group meet in this area?
2. Second veterans'' group
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3234/vshow}{GSS Data Explorer page for \code{mtvet2}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no answer \tab Total \cr
   2004 \tab 2796 \tab 9 \tab 7 \tab 2812 \cr
   Total \tab 2796 \tab 9 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtvet3}},
\code{\link{mtvet4}},
\code{\link{mtvet5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
