% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_359.R
\name{mtvet3}
\alias{mtvet3}
\title{Veteran group3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtvet3
}
\details{
Question C. Does the veterans'' group meet in this area?
3. Third veterans'' group
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3235/vshow}{GSS Data Explorer page for \code{mtvet3}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no answer \tab Total \cr
   2004 \tab 2802 \tab 3 \tab 7 \tab 2812 \cr
   Total \tab 2802 \tab 3 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtvet2}},
\code{\link{mtvet4}},
\code{\link{mtvet5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
