% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_359.R
\name{mtvet5}
\alias{mtvet5}
\title{Veteran group5 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtvet5
}
\details{
Question C. Does the veterans'' group meet in this area?
5. Fifth veterans'' group
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3237/vshow}{GSS Data Explorer page for \code{mtvet5}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtvet2}},
\code{\link{mtvet3}},
\code{\link{mtvet4}}

Other Voluntary Associations: 
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtvet2}},
\code{\link{mtvet3}},
\code{\link{mtvet4}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
