% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_265.R
\name{music30}
\alias{music30}
\title{R use music site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
music30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
J. Music/concerts?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} 3-5 times
\item \code{4} more than 5 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2403/vshow}{GSS Data Explorer page for \code{music30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 365 \tab 171 \tab 67 \tab 65 \tab 74 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 520 \tab 234 \tab 114 \tab 121 \tab 94 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 391 \tab 177 \tab 54 \tab 56 \tab 12 \tab - \tab 2812 \cr
   Total \tab 5878 \tab 1276 \tab 582 \tab 235 \tab 242 \tab 180 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{art30}},
\code{\link{games30}},
\code{\link{health30}},
\code{\link{humor30}},
\code{\link{pol30}},
\code{\link{relig30}},
\code{\link{sports30}},
\code{\link{travel30}},
\code{\link{tvmov30}}
}
\concept{Information Society}
\keyword{variable}
