% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_271.R
\name{musicbuy}
\alias{musicbuy}
\title{Use web to buy music after hearing on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
musicbuy
}
\details{
Question E. Purchased a recording as a result of hearing it on the Web?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2460/vshow}{GSS Data Explorer page for \code{musicbuy}}.
}

\seealso{
Other Information Society: 
\code{\link{musiclst}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwgthic}},
\code{\link{wwwjazz}},
\code{\link{wwwoldie}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}}
}
\concept{Information Society}
\keyword{variable}
