% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{musicdif}
\alias{musicdif}
\title{Visit many different types of music}
\source{
General Social Survey https://gss.norc.org
}
\description{
musicdif
}
\details{
Question Is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that:
B. You visit music sites that contain information about or sounds from many different types of music?
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2471/vshow}{GSS Data Explorer page for \code{musicdif}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not true \tab no answer \tab Total \cr
   2000 \tab 2498 \tab 78 \tab 81 \tab 45 \tab 115 \tab 2817 \cr
   2002 \tab 2492 \tab 103 \tab 73 \tab 51 \tab 46 \tab 2765 \cr
   Total \tab 4990 \tab 181 \tab 154 \tab 96 \tab 161 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musiclke}},
\code{\link{musicnew}},
\code{\link{wwwart}},
\code{\link{wwwworld}}
}
\concept{Information Society}
\keyword{variable}
