% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{musicinf}
\alias{musicinf}
\title{Use web to get music information}
\source{
General Social Survey https://gss.norc.org
}
\description{
musicinf
}
\details{
Question B. Used the Web to get information about a band, individual performer, orchestra, or other group?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2457/vshow}{GSS Data Explorer page for \code{musicinf}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2498 \tab 175 \tab 30 \tab 114 \tab 2817 \cr
   2002 \tab 2492 \tab 200 \tab 27 \tab 46 \tab 2765 \cr
   Total \tab 4990 \tab 375 \tab 57 \tab 160 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
