% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_271.R
\name{musiclst}
\alias{musiclst}
\title{Use web to listen to radio}
\source{
General Social Survey https://gss.norc.org
}
\description{
musiclst
}
\details{
Question D. Used the Web to listen to an online radio broadcast or view a music video?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2459/vshow}{GSS Data Explorer page for \code{musiclst}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2498 \tab 120 \tab 86 \tab 113 \tab 2817 \cr
   2002 \tab 2492 \tab 161 \tab 66 \tab 46 \tab 2765 \cr
   Total \tab 4990 \tab 281 \tab 152 \tab 159 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{musicbuy}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwgthic}},
\code{\link{wwwjazz}},
\code{\link{wwwoldie}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}}
}
\concept{Information Society}
\keyword{variable}
