% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{mustdoc}
\alias{mustdoc}
\title{X should be forced to be examined at a clinic by law}
\source{
General Social Survey https://gss.norc.org
}
\description{
mustdoc
}
\details{
Question Some cities and states have laws that force people with problems like NAME into treatment. Do you think that people like NAME should be forced by law to...
A. Get treatment at a clinic or from a doctor.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2031/vshow}{GSS Data Explorer page for \code{mustdoc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 503 \tab 844 \tab 79 \tab 18 \tab 2904 \cr
   2006 \tab 3073 \tab 413 \tab 956 \tab 52 \tab 16 \tab 4510 \cr
   2018 \tab 1175 \tab 422 \tab 693 \tab 30 \tab 28 \tab 2348 \cr
   Total \tab 5708 \tab 1338 \tab 2493 \tab 161 \tab 62 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mntlas1}},
\code{\link{musthosp}},
\code{\link{mustmed}},
\code{\link{spmentl}}
}
\concept{Mental Health}
\keyword{variable}
