% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{mustdoc1}
\alias{mustdoc1}
\title{X's parents should be forced by law to have x examined at a clinic}
\source{
General Social Survey https://gss.norc.org
}
\description{
mustdoc1
}
\details{
Question Do you think that parents of children like (VIGNAME) should be forced by law to have (VIGNAME) treated at a clinic or by a doctor?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8197/vshow}{GSS Data Explorer page for \code{mustdoc1}}.
}

\seealso{
Other Mental Health: 
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{imprvown1}},
\code{\link{morextrm}},
\code{\link{musthosp1}},
\code{\link{mustmed1}},
\code{\link{physill1}},
\code{\link{smedia}}
}
\concept{Mental Health}
\keyword{variable}
