% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{musthosp1}
\alias{musthosp1}
\title{X's parents should be forced by law to admit x to be hospitalized for treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
musthosp1
}
\details{
Question Do you think that parents of children like (VIGNAME) should be forced by law to   admit (VIGNAME)  to a hospital for treatment?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8198/vshow}{GSS Data Explorer page for \code{musthosp1}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 194 \tab 1377 \tab 41 \tab 1661 \tab 9 \tab 27 \tab 3309 \cr
   Total \tab 0 \tab 194 \tab 1377 \tab 41 \tab 1661 \tab 9 \tab 27 \tab 3309 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{bullied}},
\code{\link{imprvown1}},
\code{\link{morextrm}},
\code{\link{mustdoc1}},
\code{\link{mustmed1}},
\code{\link{smedia}}
}
\concept{Mental Health}
\keyword{variable}
