% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{myfaith}
\alias{myfaith}
\title{Degree to which faith is free of doubt}
\source{
General Social Survey https://gss.norc.org
}
\description{
myfaith
}
\details{
Question Here is another card with contrasting ideas. If you feel that
"My faith is completely free of doubts," you would place yourself at 1. If you feel "My faith is mixed with doubts," you would place yourself at 7. If you feel somewhere between these two, you would place yourself at 2, 3, 4, 5 or 6. Where would you place your feelings about your faith…
A. From Free of Doubt to Mixed with doubts
}
\section{Values}{

\itemize{
\item \verb{[1]} faith free of doubt
\item \verb{[7]} faith mixed w doubt
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1093/vshow}{GSS Data Explorer page for \code{myfaith}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab faith free of doubt \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab faith mixed w doubt \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 401 \tab 267 \tab 230 \tab 241 \tab 138 \tab 73 \tab 105 \tab 7 \tab 19 \tab 1481 \cr
   Total \tab 0 \tab 401 \tab 267 \tab 230 \tab 241 \tab 138 \tab 73 \tab 105 \tab 7 \tab 19 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{madatgod}}
}
\concept{Religion}
\keyword{variable}
