% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{mykids}
\alias{mykids}
\title{R's current insurance plan covers children}
\source{
General Social Survey https://gss.norc.org
}
\description{
mykids
}
\details{
Question Are the following people in your family covered by your current insurance plan?
B. Children for whom I am a biological or adoptive parent
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} not applicable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3933/vshow}{GSS Data Explorer page for \code{mykids}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab not applicable \tab no answer \tab Total \cr
   2008 \tab 1888 \tab 18 \tab 20 \tab 54 \tab 43 \tab 2023 \cr
   Total \tab 1888 \tab 18 \tab 20 \tab 54 \tab 43 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{othkids}},
\code{\link{sppart}}
}
\concept{Sexual Orientation}
\keyword{variable}
