% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_223.R
\name{myprobs1}
\alias{myprobs1}
\title{How much person 1's mental health condition causes r problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
myprobs1
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very much, how much does (NAME 1)'s mental health condition cause you problems, create stress, or make your life difficult?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - not at all
\item \verb{[10]} 10 - very much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7125/vshow}{GSS Data Explorer page for \code{myprobs1}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all \tab 10 - very much \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1751 \tab 275 \tab 34 \tab 61 \tab 49 \tab 26 \tab 46 \tab 25 \tab 32 \tab 32 \tab 13 \tab 3 \tab 1 \tab 2348 \cr
   Total \tab 1751 \tab 275 \tab 34 \tab 61 \tab 49 \tab 26 \tab 46 \tab 25 \tab 32 \tab 32 \tab 13 \tab 3 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
