% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_224.R
\name{myprobs4}
\alias{myprobs4}
\title{How much person 4's mental health condition causes r problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
myprobs4
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very much, how much does (NAME 4)'s mental health condition cause you problems, create stress, or make your life difficult?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - not at all
\item \verb{[10]} 10 - very much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7128/vshow}{GSS Data Explorer page for \code{myprobs4}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all \tab 10 - very much \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2190 \tab 74 \tab 2 \tab 20 \tab 10 \tab 16 \tab 9 \tab 4 \tab 4 \tab 5 \tab 4 \tab 4 \tab 6 \tab 2348 \cr
   Total \tab 2190 \tab 74 \tab 2 \tab 20 \tab 10 \tab 16 \tab 9 \tab 4 \tab 4 \tab 5 \tab 4 \tab 4 \tab 6 \tab 2348 \cr
}
}

\keyword{variable}
