% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{myskills}
\alias{myskills}
\title{Job allows r use of skills}
\source{
General Social Survey https://gss.norc.org
}
\description{
myskills
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
H. My job lets me use my skills and abilities
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2781/vshow}{GSS Data Explorer page for \code{myskills}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 711 \tab 879 \tab 142 \tab 43 \tab 7 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 698 \tab 831 \tab 155 \tab 29 \tab 9 \tab 12 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 481 \tab 576 \tab 87 \tab 19 \tab 1 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1296 \tab 477 \tab 670 \tab 82 \tab 13 \tab - \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 510 \tab 801 \tab 86 \tab 18 \tab 3 \tab 30 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 797 \tab 1002 \tab 132 \tab 29 \tab 8 \tab 13 \tab 17 \tab 3544 \cr
   Total \tab 8344 \tab 3674 \tab 4759 \tab 684 \tab 151 \tab 28 \tab 92 \tab 17 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
