% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{nafta2}
\alias{nafta2}
\title{America benefits from being a member of nafta?}
\source{
General Social Survey https://gss.norc.org
}
\description{
nafta2
}
\details{
Question Generally speaking, would you say that America benefits or does not benefit from being a member of NAFTA?
}
\section{Values}{

\itemize{
\item \verb{[1]} benefits
\item \verb{[2]} does not benefit
\item \verb{[3]} dont know
\item \verb{[4]} have never heard of nafta
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4877/vshow}{GSS Data Explorer page for \code{nafta2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab benefits \tab does not benefit \tab dont know \tab have never heard of nafta \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 377 \tab 270 \tab 581 \tab 120 \tab 19 \tab 2904 \cr
   Total \tab 1537 \tab 377 \tab 270 \tab 581 \tab 120 \tab 19 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{ethclose}},
\code{\link{nafta1}},
\code{\link{parcit}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}}
}
\concept{ISSP National Identity}
\keyword{variable}
