% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{nafta3}
\alias{nafta3}
\title{America should follow nafta decision}
\source{
General Social Survey https://gss.norc.org
}
\description{
nafta3
}
\details{
Question B. How strongly do you agree or disagree with the following statement? (Please, select one response)
America should follow NAFTA decisions, even if it does not agree with them.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4897/vshow}{GSS Data Explorer page for \code{nafta3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 22 \tab 163 \tab 319 \tab 297 \tab 64 \tab 338 \tab 13 \tab 2812 \cr
   Total \tab 1596 \tab 22 \tab 163 \tab 319 \tab 297 \tab 64 \tab 338 \tab 13 \tab 2812 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta2a}}
}
\concept{ISSP National Identity}
\keyword{variable}
