% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{nanoben}
\alias{nanoben}
\title{Balance will be strongly in favor of the benefits or slightly}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanoben
}
\details{
Question B. (IF NANOWILL=1) Would you say that the balance will be strongly in favor of the benefits, or only slightly?
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor
\item \code{2} slightly in favor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3449/vshow}{GSS Data Explorer page for \code{nanoben}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab strongly in favor \tab slightly in favor \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1446 \tab 332 \tab 224 \tab 20 \tab 1 \tab 2023 \cr
   2010 \tab 1869 \tab 98 \tab 58 \tab 7 \tab 12 \tab 2044 \cr
   2016 \tab 2411 \tab 263 \tab 177 \tab 6 \tab 10 \tab 2867 \cr
   2018 \tab 1917 \tab 235 \tab 181 \tab 5 \tab 10 \tab 2348 \cr
   Total \tab 7643 \tab 928 \tab 640 \tab 38 \tab 33 \tab 9282 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoharm}},
\code{\link{scimode}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
