% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{nanoknw1}
\alias{nanoknw1}
\title{True-false: nanotechnology manipulate extremely small units}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanoknw1
}
\details{
Question Here are a couple of true-false questions about nanotechnology. As before, if you don’t know or aren’t sure, just tell me so, and we will skip to the next question. Remember true, false, or don’t know.
A. Nanotechnology involves manipulating extremely small units of matter, such as individual atoms, in order to produce better materials. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3446/vshow}{GSS Data Explorer page for \code{nanoknw1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3688 \tab 483 \tab 63 \tab 274 \tab 2 \tab 4510 \cr
   2008 \tab 1312 \tab 449 \tab 58 \tab 200 \tab 4 \tab 2023 \cr
   2010 \tab 1796 \tab 136 \tab 22 \tab 82 \tab 8 \tab 2044 \cr
   Total \tab 6796 \tab 1068 \tab 143 \tab 556 \tab 14 \tab 8577 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
