% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_317.R
\name{natactive}
\alias{natactive}
\title{Natural environments are a good place for physical activity, such as walking, cy}
\source{
General Social Survey https://gss.norc.org
}
\description{
natactive
}
\details{
Question Thinking about your day to day experience of nature, please tell me how much do you agree or disagree with the following statements. Natural environments are a good place for physical activity, such as walking, cycling, sports, etc.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} somewhat agree
\item \verb{[3]} somewhat disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7190/vshow}{GSS Data Explorer page for \code{natactive}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 1038 \tab 102 \tab 1 \tab 5 \tab 4 \tab 23 \tab 2348 \cr
   Total \tab 1175 \tab 1038 \tab 102 \tab 1 \tab 5 \tab 4 \tab 23 \tab 2348 \cr
}
}

\keyword{variable}
