% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{nataidy}
\alias{nataidy}
\title{Assistance to other countries:ver y}
\source{
General Social Survey https://gss.norc.org
}
\description{
nataidy
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
J. Assistance to other countries
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/208/vshow}{GSS Data Explorer page for \code{nataidy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1984 \tab 974 \tab 18 \tab 79 \tab 377 \tab 14 \tab 11 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 54 \tab 158 \tab 535 \tab 29 \tab 7 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 33 \tab 120 \tab 553 \tab 28 \tab 6 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 67 \tab 235 \tab 856 \tab 42 \tab 16 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 34 \tab 142 \tab 548 \tab 26 \tab 13 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 42 \tab 157 \tab 530 \tab 31 \tab 9 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 38 \tab 138 \tab 482 \tab 24 \tab 16 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 24 \tab 152 \tab 538 \tab 33 \tab 14 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 26 \tab 156 \tab 593 \tab 19 \tab 13 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 58 \tab 254 \tab 1080 \tab 52 \tab 40 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 55 \tab 259 \tab 1043 \tab 49 \tab 57 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 88 \tab 292 \tab 978 \tab 78 \tab 15 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 109 \tab 327 \tab 898 \tab 71 \tab 3 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 123 \tab 276 \tab 966 \tab 38 \tab 4 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 113 \tab 302 \tab 947 \tab 46 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 145 \tab 317 \tab 989 \tab 55 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 1004 \tab 94 \tab 213 \tab 677 \tab 33 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 72 \tab 269 \tab 657 \tab 31 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 56 \tab 186 \tab 692 \tab 40 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 68 \tab 250 \tab 910 \tab 35 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 104 \tab 330 \tab 954 \tab 40 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 149 \tab 305 \tab 653 \tab 51 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 202 \tab 609 \tab 1220 \tab 4 \tab - \tab 10 \tab 4032 \cr
   2022 \tab 1780 \tab 172 \tab 480 \tab 1064 \tab 45 \tab - \tab 3 \tab 3544 \cr
   Total \tab 28955 \tab 1944 \tab 6006 \tab 18740 \tab 914 \tab 239 \tab 13 \tab 56811 \cr
}
}

\seealso{
Other Core: 
\code{\link{natarmsy}},
\code{\link{natcityy}},
\code{\link{natcrimy}},
\code{\link{natdrugy}},
\code{\link{nateducy}},
\code{\link{natfarey}},
\code{\link{nathealy}},
\code{\link{natracey}},
\code{\link{natspacz}}
}
\concept{Core}
\keyword{variable}
