% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{natarmsy}
\alias{natarmsy}
\title{National defense:version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
natarmsy
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
I. National defense
}
\section{Values}{

\itemize{
\item \verb{[1]} too little
\item \verb{[2]} about right
\item \verb{[3]} too much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/207/vshow}{GSS Data Explorer page for \code{natarmsy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1984 \tab 974 \tab 87 \tab 192 \tab 185 \tab 23 \tab 12 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 123 \tab 302 \tab 320 \tab 36 \tab 2 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 119 \tab 332 \tab 254 \tab 29 \tab 6 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 211 \tab 475 \tab 454 \tab 63 \tab 13 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 134 \tab 291 \tab 277 \tab 51 \tab 10 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 116 \tab 312 \tab 287 \tab 45 \tab 9 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 65 \tab 294 \tab 297 \tab 27 \tab 15 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 96 \tab 425 \tab 195 \tab 30 \tab 15 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 94 \tab 315 \tab 355 \tab 31 \tab 12 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 233 \tab 676 \tab 464 \tab 73 \tab 38 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 231 \tab 648 \tab 438 \tab 86 \tab 60 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 254 \tab 667 \tab 436 \tab 85 \tab 9 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 351 \tab 618 \tab 348 \tab 87 \tab 4 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 477 \tab 591 \tab 280 \tab 57 \tab 2 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 445 \tab 538 \tab 388 \tab 37 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 413 \tab 478 \tab 554 \tab 61 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 1004 \tab 268 \tab 345 \tab 367 \tab 36 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 243 \tab 408 \tab 330 \tab 49 \tab - \tab - \tab 2044 \cr
   2012 \tab 998 \tab 241 \tab 395 \tab 298 \tab 41 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 392 \tab 436 \tab 381 \tab 53 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 525 \tab 487 \tab 365 \tab 53 \tab - \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 341 \tab 449 \tab 325 \tab 42 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 512 \tab 915 \tab 606 \tab 5 \tab - \tab 7 \tab 4032 \cr
   2022 \tab 1780 \tab 550 \tab 684 \tab 476 \tab 50 \tab 1 \tab 3 \tab 3544 \cr
   Total \tab 28955 \tab 6521 \tab 11273 \tab 8680 \tab 1150 \tab 222 \tab 10 \tab 56811 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataidy}},
\code{\link{natcityy}},
\code{\link{natcrimy}},
\code{\link{natdrugy}},
\code{\link{nateducy}},
\code{\link{natfarey}},
\code{\link{nathealy}},
\code{\link{natracey}},
\code{\link{natspacz}}
}
\concept{Core}
\keyword{variable}
