% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_27.R
\name{natarmsz}
\alias{natarmsz}
\title{Strengthening national defense:ver z}
\source{
General Social Survey https://gss.norc.org
}
\description{
natarmsz
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
I. Strengthening national defense
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/218/vshow}{GSS Data Explorer page for \code{natarmsz}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 101 \tab 200 \tab 165 \tab 15 \tab 3 \tab 1473 \cr
   Total \tab 989 \tab 101 \tab 200 \tab 165 \tab 15 \tab 3 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataidz}},
\code{\link{natcityz}},
\code{\link{natcrimz}},
\code{\link{natdrugz}},
\code{\link{nateducz}},
\code{\link{natenviz}},
\code{\link{natfarez}},
\code{\link{nathealz}},
\code{\link{natracez}}
}
\concept{Core}
\keyword{variable}
