% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{natarts}
\alias{natarts}
\title{Spending right amount on arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
natarts
}
\details{
Question Earlier we talked about problems in this country, and whether we are spending too much money on them, too little money, or about the right amount. I have one more problem to ask you about: the arts . . . are we spending too much, too little, or about the right amount on that?
}
\section{Values}{

\itemize{
\item \verb{[1]} too little
\item \verb{[2]} about right
\item \verb{[3]} too much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2123/vshow}{GSS Data Explorer page for \code{natarts}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 303 \tab 631 \tab 269 \tab 229 \tab 13 \tab 2832 \cr
   Total \tab 1387 \tab 303 \tab 631 \tab 269 \tab 229 \tab 13 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion and Culture}
\keyword{variable}
