% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natcity}
\alias{natcity}
\title{Solving problems of big cities}
\source{
General Social Survey https://gss.norc.org
}
\description{
natcity
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
D. Solving the problems of the big cities
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/184/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 412 \tab 180 \tab 5 \tab 723 \tab 184 \tab - \tab 1504 \cr
   1974 \tab - \tab 360 \tab 216 \tab 10 \tab 736 \tab 162 \tab - \tab 1484 \cr
   1975 \tab - \tab 369 \tab 238 \tab 11 \tab 697 \tab 175 \tab - \tab 1490 \cr
   1976 \tab - \tab 391 \tab 174 \tab 7 \tab 636 \tab 291 \tab - \tab 1499 \cr
   1977 \tab - \tab 400 \tab 214 \tab 5 \tab 615 \tab 296 \tab - \tab 1530 \cr
   1978 \tab - \tab 452 \tab 197 \tab 1 \tab 593 \tab 289 \tab - \tab 1532 \cr
   1980 \tab - \tab 381 \tab 186 \tab 4 \tab 585 \tab 312 \tab - \tab 1468 \cr
   1982 \tab - \tab 412 \tab 240 \tab 6 \tab 870 \tab 332 \tab - \tab 1860 \cr
   1983 \tab - \tab 468 \tab 218 \tab 8 \tab 658 \tab 247 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 148 \tab 63 \tab 5 \tab 214 \tab 60 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 244 \tab 86 \tab 4 \tab 294 \tab 123 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 221 \tab 69 \tab 10 \tab 314 \tab 116 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 184 \tab 90 \tab 3 \tab 254 \tab 72 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 209 \tab 105 \tab 5 \tab 327 \tab 72 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 222 \tab 105 \tab 4 \tab 354 \tab 83 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 178 \tab 87 \tab 5 \tab 337 \tab 67 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 231 \tab 85 \tab 7 \tab 348 \tab 85 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 175 \tab 96 \tab 7 \tab 429 \tab 92 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 343 \tab 120 \tab 4 \tab 844 \tab 197 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 336 \tab 163 \tab 4 \tab 752 \tab 186 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 412 \tab 170 \tab 3 \tab 629 \tab 167 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 439 \tab 190 \tab 7 \tab 627 \tab 146 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 494 \tab 116 \tab 3 \tab 562 \tab 183 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 522 \tab 149 \tab 1 \tab 556 \tab 173 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 510 \tab 135 \tab - \tab 666 \tab 173 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 333 \tab 114 \tab - \tab 444 \tab 113 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 337 \tab 104 \tab 1 \tab 393 \tab 179 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 339 \tab 109 \tab 8 \tab 385 \tab 157 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 419 \tab 124 \tab 1 \tab 544 \tab 186 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 455 \tab 118 \tab 4 \tab 683 \tab 177 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 365 \tab 127 \tab 2 \tab 573 \tab 121 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 591 \tab 18 \tab - \tab 1009 \tab 349 \tab 20 \tab 4032 \cr
   2022 \tab 1764 \tab 443 \tab 88 \tab 5 \tab 1003 \tab 231 \tab 10 \tab 3544 \cr
   2024 \tab 1626 \tab 410 \tab 84 \tab 2 \tab 921 \tab 260 \tab 6 \tab 3309 \cr
   Total \tab 33097 \tab 12205 \tab 4578 \tab 152 \tab 19575 \tab 6056 \tab 36 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}

Other Federal Governmentending: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}

Other Split Ballots: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}

Other United States: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}
}
\concept{Core}
\concept{Federal Governmentending}
\concept{Split Ballots}
\concept{United States}
\keyword{variable}
