% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{natcityy}
\alias{natcityy}
\title{Assistance to big cities:version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
natcityy
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
D. Assistance to big cities
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/202/vshow}{GSS Data Explorer page for \code{natcityy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1984 \tab 974 \tab 89 \tab 178 \tab 161 \tab 57 \tab 14 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 139 \tab 258 \tab 289 \tab 95 \tab 2 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 112 \tab 279 \tab 241 \tab 100 \tab 8 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 287 \tab 419 \tab 327 \tab 170 \tab 13 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 161 \tab 258 \tab 213 \tab 121 \tab 10 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 145 \tab 254 \tab 244 \tab 118 \tab 8 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 153 \tab 250 \tab 185 \tab 96 \tab 14 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 151 \tab 247 \tab 226 \tab 117 \tab 20 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 187 \tab 242 \tab 246 \tab 120 \tab 12 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 347 \tab 471 \tab 432 \tab 199 \tab 35 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 311 \tab 451 \tab 417 \tab 224 \tab 60 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 275 \tab 524 \tab 398 \tab 242 \tab 12 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 324 \tab 483 \tab 357 \tab 238 \tab 6 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 221 \tab 579 \tab 453 \tab 148 \tab 6 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 242 \tab 580 \tab 402 \tab 183 \tab 4 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 290 \tab 567 \tab 471 \tab 178 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 1005 \tab 192 \tab 394 \tab 286 \tab 144 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 177 \tab 381 \tab 321 \tab 147 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 140 \tab 341 \tab 322 \tab 167 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 222 \tab 464 \tab 428 \tab 148 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 309 \tab 529 \tab 439 \tab 152 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 242 \tab 433 \tab 344 \tab 139 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 526 \tab 846 \tab 642 \tab 9 \tab 1 \tab 21 \tab 4032 \cr
   2022 \tab 1780 \tab 423 \tab 643 \tab 574 \tab 117 \tab - \tab 7 \tab 3544 \cr
   Total \tab 28956 \tab 5665 \tab 10071 \tab 8418 \tab 3429 \tab 244 \tab 28 \tab 56811 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataidy}},
\code{\link{natarmsy}},
\code{\link{natcrimy}},
\code{\link{natdrugy}},
\code{\link{nateducy}},
\code{\link{natfarey}},
\code{\link{nathealy}},
\code{\link{natracey}},
\code{\link{natspacz}}
}
\concept{Core}
\keyword{variable}
