% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{natdeg}
\alias{natdeg}
\title{In which country did you obtain your highest degree?}
\source{
General Social Survey https://gss.norc.org
}
\description{
natdeg
}
\details{
Question In which country did you obtain your highest degree?
}
\section{Values}{

\itemize{
\item \verb{[1]} usa
\item \verb{[2]} antigua
\item \verb{[3]} trinidad
\item \verb{[4]} jamaica
\item \verb{[5]} pakistan
\item \verb{[6]} turkey
\item \verb{[7]} italy
\item \verb{[8]} west indies
\item \verb{[9]} colombia
\item \verb{[10]} kazakhstan
\item \verb{[12]} cuba
\item \verb{[13]} india
\item \verb{[14]} germany
\item \verb{[15]} ecuador
\item \verb{[16]} mexico
\item \verb{[17]} portugal
\item \verb{[18]} el salvador
\item \verb{[19]} canada
\item \verb{[20]} guatemala
\item \verb{[21]} zimbabwe
\item \verb{[23]} taiwan
\item \verb{[24]} philippines
\item \verb{[25]} israel
\item \verb{[26]} dominican republic
\item \verb{[27]} scotland
\item \verb{[28]} haiti
\item \verb{[29]} honduras
\item \verb{[30]} puerto rico
\item \verb{[31]} austraila
\item \verb{[32]} china
\item \verb{[33]} poland
\item \verb{[34]} lithuania
\item \verb{[35]} liberia
\item \verb{[36]} brazil
\item \verb{[37]} morocco
\item \verb{[38]} spain
\item \verb{[39]} bulgeria
\item \verb{[40]} ussr
\item \verb{[41]} kuwait
\item \verb{[42]} japan
\item \verb{[43]} korea
\item \verb{[44]} vietnam
\item \verb{[45]} bermuda
\item \verb{[46]} ghana
\item \verb{[47]} panama
\item \verb{[48]} france
\item \verb{[49]} romania
\item \verb{[50]} south africa
\item \verb{[51]} england
\item \verb{[52]} britain
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3902/vshow}{GSS Data Explorer page for \code{natdeg}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab usa \tab trinidad \tab jamaica \tab pakistan \tab turkey \tab west indies \tab colombia \tab kazakhstan \tab cuba \tab india \tab germany \tab ecuador \tab mexico \tab el salvador \tab canada \tab guatemala \tab philippines \tab israel \tab dominican republic \tab scotland \tab haiti \tab honduras \tab puerto rico \tab austraila \tab china \tab poland \tab lithuania \tab brazil \tab morocco \tab spain \tab bulgeria \tab ussr \tab japan \tab korea \tab bermuda \tab france \tab south africa \tab england \tab britain \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 603 \tab 1239 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab 6 \tab 2 \tab 4 \tab 7 \tab 2 \tab 1 \tab 39 \tab 7 \tab 3 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 60 \tab - \tab 2023 \cr
   Total \tab 603 \tab 1239 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab 6 \tab 2 \tab 4 \tab 7 \tab 2 \tab 1 \tab 39 \tab 7 \tab 3 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 60 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{autonojb}},
\code{\link{bornsp}},
\code{\link{curempyr}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{typevisa}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work3yrs}}

Other Education: 
\code{\link{actgreek}},
\code{\link{actschl}},
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{bilinged}},
\code{\link{cathsch}},
\code{\link{chldeduc}},
\code{\link{chrissch}},
\code{\link{classics}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{coldeg1}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{colop}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{datesch}},
\code{\link{degree}},
\code{\link{denomsch}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{educ}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{educemp}},
\code{\link{educop}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{govedop}},
\code{\link{homesch}},
\code{\link{hsbasics}},
\code{\link{hsbio}},
\code{\link{hscaring}},
\code{\link{hschem}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsmath}},
\code{\link{hsorder}},
\code{\link{hsphys}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inteduc}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{majorcol}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memschl}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{opeduc}},
\code{\link{oppared}},
\code{\link{othsch}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{relsch}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{schleth}},
\code{\link{schlhead}},
\code{\link{schoolng}},
\code{\link{scimath}},
\code{\link{sexeduc}},
\code{\link{solgreek}},
\code{\link{solschl}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{speduc}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{teacheth}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{trstprof}},
\code{\link{undoccol}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{whoteach}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Education}
\concept{Global Economics}
\concept{Immigration}
\keyword{variable}
