% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_27.R
\name{nateducz}
\alias{nateducz}
\title{Improving the education system:ver z}
\source{
General Social Survey https://gss.norc.org
}
\description{
nateducz
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
G. Improving the nation's education system
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/216/vshow}{GSS Data Explorer page for \code{nateducz}}.
}

\seealso{
Other Core: 
\code{\link{nataidz}},
\code{\link{natarmsz}},
\code{\link{natcityz}},
\code{\link{natcrimz}},
\code{\link{natdrugz}},
\code{\link{natenviz}},
\code{\link{natfarez}},
\code{\link{nathealz}},
\code{\link{natracez}}

Other Federal Government: 
\code{\link{nataidz}},
\code{\link{natarmsz}},
\code{\link{natcityz}},
\code{\link{natcrimz}},
\code{\link{natdrugz}},
\code{\link{natenviz}},
\code{\link{natfarez}},
\code{\link{nathealz}},
\code{\link{natracez}}

Other Spending: 
\code{\link{nataidz}},
\code{\link{natarmsz}},
\code{\link{natcityz}},
\code{\link{natcrimz}},
\code{\link{natdrugz}},
\code{\link{natenviz}},
\code{\link{natfarez}},
\code{\link{nathealz}},
\code{\link{natracez}}

Other Split Ballots: 
\code{\link{nataidz}},
\code{\link{natarmsz}},
\code{\link{natcityz}},
\code{\link{natcrimz}},
\code{\link{natenviz}},
\code{\link{natfarez}},
\code{\link{nathealz}},
\code{\link{natracez}}

Other United States: 
\code{\link{nataidz}},
\code{\link{natarmsz}},
\code{\link{natcityz}},
\code{\link{natcrimz}},
\code{\link{natenviz}},
\code{\link{natfarez}},
\code{\link{nathealz}},
\code{\link{natracez}}
}
\concept{Core}
\concept{Education}
\concept{Federal Government}
\concept{Spending}
\concept{Split Ballots}
\concept{United States}
\keyword{variable}
