% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_25.R
\name{natenrgy}
\alias{natenrgy}
\title{Developing alternative energy sources}
\source{
General Social Survey https://gss.norc.org
}
\description{
natenrgy
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
R. Developing alternative energy sources
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/198/vshow}{GSS Data Explorer page for \code{natenrgy}}.
}

\seealso{
Other Core: 
\code{\link{natchld}},
\code{\link{natenviy}},
\code{\link{natfare}},
\code{\link{natmass}},
\code{\link{natpark}},
\code{\link{natroad}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natspacy}}

Other Environment: 
\code{\link{natenviy}}

Other Split Ballots: 
\code{\link{natchld}},
\code{\link{natenviy}},
\code{\link{natfare}},
\code{\link{natmass}},
\code{\link{natpark}},
\code{\link{natroad}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natspacy}}

Other United States: 
\code{\link{natchld}},
\code{\link{natenviy}},
\code{\link{natfare}},
\code{\link{natmass}},
\code{\link{natpark}},
\code{\link{natroad}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natspacy}}
}
\concept{Core}
\concept{Environment}
\concept{Split Ballots}
\concept{United States}
\keyword{variable}
