% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{nathealy}
\alias{nathealy}
\title{Health:version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
nathealy
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
C. Health
}
\section{Values}{

\itemize{
\item \verb{[1]} too little
\item \verb{[2]} about right
\item \verb{[3]} too much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/201/vshow}{GSS Data Explorer page for \code{nathealy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1984 \tab 974 \tab 270 \tab 158 \tab 41 \tab 19 \tab 11 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 432 \tab 253 \tab 76 \tab 19 \tab 3 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 431 \tab 226 \tab 50 \tab 23 \tab 10 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 806 \tab 296 \tab 66 \tab 36 \tab 12 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 514 \tab 179 \tab 39 \tab 22 \tab 9 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 536 \tab 171 \tab 41 \tab 12 \tab 9 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 457 \tab 168 \tab 37 \tab 21 \tab 15 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 509 \tab 189 \tab 29 \tab 18 \tab 16 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 555 \tab 99 \tab 106 \tab 32 \tab 15 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 906 \tab 294 \tab 188 \tab 58 \tab 38 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 879 \tab 341 \tab 152 \tab 34 \tab 57 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 975 \tab 305 \tab 109 \tab 51 \tab 11 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 976 \tab 306 \tab 85 \tab 38 \tab 3 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 1026 \tab 265 \tab 95 \tab 18 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 1093 \tab 200 \tab 95 \tab 20 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 1086 \tab 258 \tab 137 \tab 27 \tab - \tab - \tab 4510 \cr
   2008 \tab 1005 \tab 740 \tab 147 \tab 114 \tab 17 \tab - \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 601 \tab 184 \tab 210 \tab 33 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 572 \tab 207 \tab 161 \tab 36 \tab - \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 701 \tab 272 \tab 254 \tab 34 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 887 \tab 303 \tab 209 \tab 31 \tab - \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 802 \tab 220 \tab 116 \tab 22 \tab - \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 1333 \tab 432 \tab 260 \tab 4 \tab - \tab 16 \tab 4032 \cr
   2022 \tab 1780 \tab 1208 \tab 330 \tab 201 \tab 22 \tab - \tab 3 \tab 3544 \cr
   Total \tab 28956 \tab 18295 \tab 5803 \tab 2871 \tab 647 \tab 220 \tab 19 \tab 56811 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataidy}},
\code{\link{natarmsy}},
\code{\link{natcityy}},
\code{\link{natcrimy}},
\code{\link{natdrugy}},
\code{\link{nateducy}},
\code{\link{natfarey}},
\code{\link{natracey}},
\code{\link{natspacz}}
}
\concept{Core}
\keyword{variable}
