% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_27.R
\name{nathealz}
\alias{nathealz}
\title{Improving health:version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
nathealz
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
C. Improving and protecting the nation's health
}
\section{Values}{

\itemize{
\item \verb{[1]} too little
\item \verb{[2]} about right
\item \verb{[3]} too much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/212/vshow}{GSS Data Explorer page for \code{nathealz}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 273 \tab 169 \tab 23 \tab 15 \tab 4 \tab 1473 \cr
   Total \tab 989 \tab 273 \tab 169 \tab 23 \tab 15 \tab 4 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataidz}},
\code{\link{natarmsz}},
\code{\link{natcityz}},
\code{\link{natcrimz}},
\code{\link{natdrugz}},
\code{\link{nateducz}},
\code{\link{natenviz}},
\code{\link{natfarez}},
\code{\link{natracez}}
}
\concept{Core}
\keyword{variable}
