% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_317.R
\name{natlack}
\alias{natlack}
\title{There is a lack of interesting activities in natural environments}
\source{
General Social Survey https://gss.norc.org
}
\description{
natlack
}
\details{
Question Thinking about your day to day experience of nature, please tell me how much do you agree or disagree with the following statements. There is a lack of interesting activities in natural environments.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7193/vshow}{GSS Data Explorer page for \code{natlack}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 98 \tab 224 \tab 375 \tab 437 \tab 16 \tab 23 \tab 2348 \cr
   Total \tab 1175 \tab 98 \tab 224 \tab 375 \tab 437 \tab 16 \tab 23 \tab 2348 \cr
}
}

\keyword{variable}
