% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{natlart}
\alias{natlart}
\title{Feel fed gov should provide assistance to art}
\source{
General Social Survey https://gss.norc.org
}
\description{
natlart
}
\details{
Question If arts organizations -- such as art museums, dance, operas, theater groups and symphony orchestras -- need financial assistance to operate, Do you feel that the federal government should provide assistance or not?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2134/vshow}{GSS Data Explorer page for \code{natlart}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{drama}},
\code{\link{localart}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{stateart}},
\code{\link{volarts}}
}
\concept{Religion and Culture}
\keyword{variable}
