% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_316.R
\name{natnotice}
\alias{natnotice}
\title{In my everyday life, i take notice of the natural elements around me (trees, wat}
\source{
General Social Survey https://gss.norc.org
}
\description{
natnotice
}
\details{
Question In my everyday life, I take notice of the natural elements around me (trees, water, wildlife, etc.).
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7184/vshow}{GSS Data Explorer page for \code{natnotice}}.
}

\keyword{variable}
