% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{natracey}
\alias{natracey}
\title{Assistance to black people: version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
natracey
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?H. Assistance to Black people.
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/206/vshow}{GSS Data Explorer page for \code{natracey}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1984 \tab 974 \tab 119 \tab 203 \tab 116 \tab 48 \tab 13 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 200 \tab 314 \tab 193 \tab 74 \tab 2 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 166 \tab 331 \tab 191 \tab 45 \tab 7 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 433 \tab 451 \tab 220 \tab 101 \tab 11 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 191 \tab 301 \tab 171 \tab 90 \tab 10 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 207 \tab 308 \tab 167 \tab 79 \tab 8 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 197 \tab 291 \tab 142 \tab 53 \tab 15 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 227 \tab 299 \tab 141 \tab 76 \tab 18 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 188 \tab 306 \tab 191 \tab 108 \tab 14 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 349 \tab 565 \tab 378 \tab 156 \tab 36 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 334 \tab 544 \tab 343 \tab 187 \tab 55 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 381 \tab 565 \tab 316 \tab 180 \tab 9 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 425 \tab 522 \tab 280 \tab 172 \tab 9 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 338 \tab 591 \tab 334 \tab 133 \tab 11 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 363 \tab 615 \tab 250 \tab 165 \tab 18 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 378 \tab 608 \tab 315 \tab 195 \tab 12 \tab - \tab 4510 \cr
   2008 \tab 1004 \tab 279 \tab 404 \tab 209 \tab 118 \tab 9 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 252 \tab 427 \tab 190 \tab 147 \tab 14 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 206 \tab 407 \tab 195 \tab 156 \tab 12 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 316 \tab 510 \tab 262 \tab 161 \tab 15 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 534 \tab 535 \tab 216 \tab 139 \tab 6 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 507 \tab 379 \tab 140 \tab 128 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 930 \tab 692 \tab 384 \tab 16 \tab 3 \tab 20 \tab 4032 \cr
   2022 \tab 1780 \tab 808 \tab 566 \tab 278 \tab 97 \tab 8 \tab 7 \tab 3544 \cr
   Total \tab 28955 \tab 8328 \tab 10734 \tab 5622 \tab 2824 \tab 321 \tab 27 \tab 56811 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataidy}},
\code{\link{natarmsy}},
\code{\link{natcityy}},
\code{\link{natcrimy}},
\code{\link{natdrugy}},
\code{\link{nateducy}},
\code{\link{natfarey}},
\code{\link{nathealy}},
\code{\link{natspacz}}
}
\concept{Core}
\keyword{variable}
