% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_25.R
\name{natsoc}
\alias{natsoc}
\title{Social security}
\source{
General Social Survey https://gss.norc.org
}
\description{
natsoc
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
M. Social Security
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/193/vshow}{GSS Data Explorer page for \code{natsoc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1984 \tab 484 \tab 493 \tab 344 \tab 85 \tab 46 \tab 21 \tab - \tab 1473 \cr
   1985 \tab - \tab 794 \tab 586 \tab 101 \tab 46 \tab 7 \tab - \tab 1534 \cr
   1986 \tab - \tab 810 \tab 519 \tab 90 \tab 37 \tab 14 \tab - \tab 1470 \cr
   1987 \tab - \tab 1046 \tab 570 \tab 110 \tab 74 \tab 19 \tab - \tab 1819 \cr
   1988 \tab - \tab 772 \tab 553 \tab 81 \tab 61 \tab 14 \tab - \tab 1481 \cr
   1989 \tab - \tab 826 \tab 562 \tab 66 \tab 73 \tab 10 \tab - \tab 1537 \cr
   1990 \tab - \tab 667 \tab 538 \tab 76 \tab 70 \tab 21 \tab - \tab 1372 \cr
   1991 \tab - \tab 774 \tab 580 \tab 60 \tab 81 \tab 22 \tab - \tab 1517 \cr
   1993 \tab - \tab 693 \tab 686 \tab 113 \tab 99 \tab 15 \tab - \tab 1606 \cr
   1994 \tab - \tab 1371 \tab 1221 \tab 205 \tab 151 \tab 44 \tab - \tab 2992 \cr
   1996 \tab - \tab 1388 \tab 1066 \tab 224 \tab 163 \tab 63 \tab - \tab 2904 \cr
   1998 \tab - \tab 1578 \tab 903 \tab 177 \tab 161 \tab 13 \tab - \tab 2832 \cr
   2000 \tab - \tab 1626 \tab 897 \tab 131 \tab 157 \tab 6 \tab - \tab 2817 \cr
   2002 \tab - \tab 1618 \tab 920 \tab 123 \tab 99 \tab 5 \tab - \tab 2765 \cr
   2004 \tab - \tab 1772 \tab 778 \tab 152 \tab 103 \tab 7 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 1835 \tab 891 \tab 138 \tab 124 \tab 4 \tab - \tab 4510 \cr
   2008 \tab - \tab 1206 \tab 608 \tab 112 \tab 93 \tab 4 \tab - \tab 2023 \cr
   2010 \tab - \tab 1126 \tab 677 \tab 145 \tab 91 \tab 5 \tab - \tab 2044 \cr
   2012 \tab - \tab 1048 \tab 683 \tab 146 \tab 94 \tab 3 \tab - \tab 1974 \cr
   2014 \tab - \tab 1353 \tab 906 \tab 152 \tab 124 \tab 3 \tab - \tab 2538 \cr
   2016 \tab - \tab 1690 \tab 922 \tab 144 \tab 108 \tab 3 \tab - \tab 2867 \cr
   2018 \tab - \tab 1381 \tab 748 \tab 101 \tab 115 \tab 3 \tab - \tab 2348 \cr
   2021 \tab - \tab 2286 \tab 1518 \tab 198 \tab 10 \tab - \tab 20 \tab 4032 \cr
   2022 \tab - \tab 2213 \tab 1051 \tab 161 \tab 107 \tab 3 \tab 9 \tab 3544 \cr
   Total \tab 2002 \tab 30366 \tab 18727 \tab 3091 \tab 2287 \tab 309 \tab 29 \tab 56811 \cr
}
}

\seealso{
Other Core: 
\code{\link{natchld}},
\code{\link{natenrgy}},
\code{\link{natenviy}},
\code{\link{natfare}},
\code{\link{natmass}},
\code{\link{natpark}},
\code{\link{natroad}},
\code{\link{natsci}},
\code{\link{natspacy}}
}
\concept{Core}
\keyword{variable}
