% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_520.R
\name{naturgod}
\alias{naturgod}
\title{Relationship of nature to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
naturgod
}
\details{
Question Please check one box to show which statement is closest to your views.
}
\section{Values}{

\itemize{
\item \code{1} nature created by god
\item \code{2} nature sacred in itself
\item \code{3} nature import, but not sacred
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4735/vshow}{GSS Data Explorer page for \code{naturgod}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grntaxes}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{resnatur}}

Other Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grntaxes}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{resnatur}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
