% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_38.R
\name{neargod}
\alias{neargod}
\title{How close does r feel to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
neargod
}
\details{
Question How close do you feel to God most of the time?  Would you say extremely close, somewhat close, not very close, or not close at all?
}
\section{Values}{

\itemize{
\item \code{1} extremely close
\item \code{2} somewhat close
\item \code{3} not very close
\item \code{4} not close at all
\item \code{5} does not believe
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/321/vshow}{GSS Data Explorer page for \code{neargod}}.
}

\seealso{
Other Core: 
\code{\link{creator}},
\code{\link{father}},
\code{\link{friend}},
\code{\link{judge}},
\code{\link{lover}},
\code{\link{master}},
\code{\link{mother}},
\code{\link{redeemer}},
\code{\link{spouse}}

Other Religion: 
\code{\link{creator}},
\code{\link{father}},
\code{\link{friend}},
\code{\link{judge}},
\code{\link{lover}},
\code{\link{master}},
\code{\link{mother}},
\code{\link{redeemer}},
\code{\link{spouse}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
