% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_209.R
\name{needynei}
\alias{needynei}
\title{Contributed to a needy neighbor.}
\source{
General Social Survey https://gss.norc.org
}
\description{
needynei
}
\details{
Question People help other people in ways besides giving money, time, or other things to organized groups. Sometimes people help needy people directly. During the past 12 months, did you or members of your family or household give money, food, or clothing to any of the following types of people?
B. A needy neighbor
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1976/vshow}{GSS Data Explorer page for \code{needynei}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 650 \tab 769 \tab 6 \tab 19 \tab 2904 \cr
   Total \tab 1460 \tab 650 \tab 769 \tab 6 \tab 19 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{homeless}},
\code{\link{needyfrd}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{valinfrm}},
\code{\link{valintl}},
\code{\link{valoth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
