% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_209.R
\name{needyrel}
\alias{needyrel}
\title{Contributed to a needy relative.}
\source{
General Social Survey https://gss.norc.org
}
\description{
needyrel
}
\details{
Question People help other people in ways besides giving money, time, or other things to organized groups. Sometimes people help needy people directly. During the past 12 months, did you or members of your family or household give money, food, or clothing to any of the following types of people?
C. A needy relative
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1977/vshow}{GSS Data Explorer page for \code{needyrel}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 364 \tab 1036 \tab 15 \tab 29 \tab 2904 \cr
   Total \tab 1460 \tab 364 \tab 1036 \tab 15 \tab 29 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{homeless}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{valinfrm}},
\code{\link{valintl}},
\code{\link{valoth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
