% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_262.R
\name{neieth14}
\alias{neieth14}
\title{Preferred ethnic neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
neieth14
}
\details{
Question Now I’d like you to imagine a neighborhood that had an ethnic and racial mix you personally would feel most comfortable in. Here is a blank neighborhood card, which depicts some houses that surround your own. Using the letters A for Asian, B for Black, H for Hispanic or Latin American and W for White, please put a letter in each of these houses to represent your preferred neighborhood where you would most like to live. Please be sure to fill in all ofthe houses.
}
\section{Values}{

\itemize{
\item \code{1} asian
\item \code{2} black
\item \code{3} hispanic
\item \code{4} white
\item \code{5} doesnt matter-dont care-race not important
\item \code{6} rural-wants no neighbors
\item \code{7} mixed race, 2 or more races in 1 hh
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2375/vshow}{GSS Data Explorer page for \code{neieth14}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}}

Other Race: 
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}}
}
\concept{Multi-Ethnic United States}
\concept{Race}
\keyword{variable}
