% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_261.R
\name{neieth5}
\alias{neieth5}
\title{Preferred ethnic neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
neieth5
}
\details{
Question Now I'd like you to imagine a neighborhood that had an ethnic and racial mix you personally would feel most comfortable in. Here is a blank neighborhood card, which depicts some houses that surround your own. Using the letters A for Asian, B for Black, H for Hispanic or Latin American and W for White, please put a letter in each of these houses to represent your preferred neighborhood where you would most like to live. Please be sure to fill in all ofthe houses.
}
\section{Values}{

\itemize{
\item \verb{[1]} asian
\item \verb{[2]} black
\item \verb{[3]} hispanic
\item \verb{[4]} white
\item \verb{[5]} doesnt matter-dont care-race not important
\item \verb{[6]} rural-wants no neighbors
\item \verb{[7]} mixed race, 2 or more races in 1 hh
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2366/vshow}{GSS Data Explorer page for \code{neieth5}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab asian \tab black \tab hispanic \tab white \tab doesnt matter-dont care-race not important \tab rural-wants no neighbors \tab mixed race, 2 or more races in 1 hh \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 196 \tab 224 \tab 165 \tab 574 \tab 47 \tab 5 \tab 5 \tab 2 \tab 180 \tab 2817 \cr
   Total \tab 1419 \tab 196 \tab 224 \tab 165 \tab 574 \tab 47 \tab 5 \tab 5 \tab 2 \tab 180 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{leastcom}},
\code{\link{neieth1}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
