% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{neifrd}
\alias{neifrd}
\title{Number of close friends who live near r}
\source{
General Social Survey https://gss.norc.org
}
\description{
neifrd
}
\details{
Question Thinking now of people who live near you - in your neighborhood or district: How many of these people are close friends of yours?
}
\section{Values}{

\itemize{
\item \code{0} no close friends living near r
\item \code{96} more than a hundred
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4203/vshow}{GSS Data Explorer page for \code{neifrd}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab no close friends living near r \tab 1 \tab 10 \tab 12 \tab 15 \tab 16 \tab 19 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 356 \tab 122 \tab 42 \tab 7 \tab 11 \tab 2 \tab 1 \tab 200 \tab 19 \tab 3 \tab 110 \tab 1 \tab 96 \tab 86 \tab 4 \tab 52 \tab 11 \tab 9 \tab 10 \tab 1 \tab 6 \tab 2765 \cr
   Total \tab 1616 \tab 356 \tab 122 \tab 42 \tab 7 \tab 11 \tab 2 \tab 1 \tab 200 \tab 19 \tab 3 \tab 110 \tab 1 \tab 96 \tab 86 \tab 4 \tab 52 \tab 11 \tab 9 \tab 10 \tab 1 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
