% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_100.R
\name{neighbr5}
\alias{neighbr5}
\title{Person #5 is r's neighbor}
\source{
General Social Survey https://gss.norc.org
}
\description{
neighbr5
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
H5. Fifth Person is Neighbor
}
\section{Values}{

\itemize{
\item \verb{[1]} mentioned
\item \verb{[2]} not mentioned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/914/vshow}{GSS Data Explorer page for \code{neighbr5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 30 \tab 348 \tab 3 \tab 1534 \cr
   2004 \tab 2600 \tab 6 \tab 160 \tab 46 \tab 2812 \cr
   Total \tab 3753 \tab 36 \tab 508 \tab 49 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}}
}
\concept{Social Networks}
\keyword{variable}
