% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{nervous}
\alias{nervous}
\title{In past 30 days feel nervous}
\source{
General Social Survey https://gss.norc.org
}
\description{
nervous
}
\details{
Question In the past 30 days, about how often did you feel:
B. Nervous.
}
\section{Values}{

\itemize{
\item \verb{[1]} all of the time
\item \verb{[2]} most of the time
\item \verb{[3]} some of the time
\item \verb{[4]} little of time
\item \verb{[5]} none of the time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2115/vshow}{GSS Data Explorer page for \code{nervous}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab all of the time \tab most of the time \tab some of the time \tab little of time \tab none of the time \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 1387 \tab 44 \tab 92 \tab 390 \tab 464 \tab 439 \tab 4 \tab 12 \tab - \tab 2832 \cr
   Total \tab 1387 \tab 44 \tab 92 \tab 390 \tab 464 \tab 439 \tab 4 \tab 12 \tab 0 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{beausprt}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{drama}},
\code{\link{effort}},
\code{\link{feelgod}},
\code{\link{fidgety}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godlove}},
\code{\link{godwatch}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{lesspain}},
\code{\link{localart}},
\code{\link{meditate}},
\code{\link{natarts}},
\code{\link{natlart}},
\code{\link{nocheer}},
\code{\link{popmusic}},
\code{\link{privpray}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{rellife}},
\code{\link{relpersn}},
\code{\link{sprtprsn}},
\code{\link{stateart}},
\code{\link{uniongod}},
\code{\link{volarts}},
\code{\link{wrthless}}

Other Feelings: 
\code{\link{accptoth}},
\code{\link{afailure}},
\code{\link{angrlast}},
\code{\link{angry}},
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{anxious}},
\code{\link{ashamed}},
\code{\link{atease}},
\code{\link{badbrks}},
\code{\link{beplesnt}},
\code{\link{big5a1}},
\code{\link{big5a2}},
\code{\link{big5b1}},
\code{\link{big5b2}},
\code{\link{big5c1}},
\code{\link{big5c2}},
\code{\link{big5d1}},
\code{\link{big5d2}},
\code{\link{big5e1}},
\code{\link{big5e2}},
\code{\link{calm}},
\code{\link{chnang1}},
\code{\link{chnang10}},
\code{\link{chnang11}},
\code{\link{chnang12}},
\code{\link{chnang13}},
\code{\link{chnang14}},
\code{\link{chnang15}},
\code{\link{chnang16}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{chnang6}},
\code{\link{chnang7}},
\code{\link{chnang8}},
\code{\link{chnang9}},
\code{\link{contentd}},
\code{\link{effort}},
\code{\link{embarrss}},
\code{\link{empathy1}},
\code{\link{empathy2}},
\code{\link{empathy3}},
\code{\link{empathy4}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{fidgety}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{hapfeel}},
\code{\link{hopeless}},
\code{\link{howangry}},
\code{\link{likeoth}},
\code{\link{litcntrl}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{madat1}},
\code{\link{madat10}},
\code{\link{madat11}},
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat19}},
\code{\link{madat2}},
\code{\link{madat20}},
\code{\link{madat21}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat5}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}},
\code{\link{moregood}},
\code{\link{mostluck}},
\code{\link{nocheer}},
\code{\link{noemote}},
\code{\link{nogood}},
\code{\link{noplan}},
\code{\link{notcount}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{othresp}},
\code{\link{outraged}},
\code{\link{ovrjoyed}},
\code{\link{pessimst}},
\code{\link{proud}},
\code{\link{reactok}},
\code{\link{restless}},
\code{\link{sad}},
\code{\link{satself}},
\code{\link{selfless}},
\code{\link{selfresp}},
\code{\link{shakeblu}},
\code{\link{showangr}},
\code{\link{showfeel}},
\code{\link{slfrspct}},
\code{\link{thnkangr}},
\code{\link{whnangry}},
\code{\link{worried}},
\code{\link{wrthless}}
}
\concept{Feelings}
\concept{Religion and Culture}
\keyword{variable}
