% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{newfrds}
\alias{newfrds}
\title{At these occasions, how often did r make new friends or acquaintences}
\source{
General Social Survey https://gss.norc.org
}
\description{
newfrds
}
\details{
Question At these occasions, how often do you make new friends or acquaintances?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} rarely
\item \code{3} sometimes
\item \code{4} often
\item \code{5} very often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7262/vshow}{GSS Data Explorer page for \code{newfrds}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab rarely \tab sometimes \tab often \tab very often \tab don't know \tab Total \cr
   2018 \tab 1404 \tab 86 \tab 296 \tab 413 \tab 95 \tab 49 \tab 5 \tab 2348 \cr
   Total \tab 1404 \tab 86 \tab 296 \tab 413 \tab 95 \tab 49 \tab 5 \tab 2348 \cr
}
}

\keyword{variable}
