% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{newjob}
\alias{newjob}
\title{How helpful present work skill would be for new job}
\source{
General Social Survey https://gss.norc.org
}
\description{
newjob
}
\details{
Question If you were to look for a new job, how helpful would your present work experience and/or job skills be?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} likely
\item \code{3} unlikely
\item \code{4} very unlikely
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4616/vshow}{GSS Data Explorer page for \code{newjob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
