% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{newjobs}
\alias{newjobs}
\title{Jobs created are better than r's current job}
\source{
General Social Survey https://gss.norc.org
}
\description{
newjobs
}
\details{
Question On average, do you think the jobs that have been created in the United States for people with jobs like yours are better than your job (with higher pay and better benefits) or worse?
}
\section{Values}{

\itemize{
\item \verb{[1]} better
\item \verb{[2]} about the same
\item \verb{[3]} worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3885/vshow}{GSS Data Explorer page for \code{newjobs}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab better \tab about the same \tab worse \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1887 \tab 44 \tab 56 \tab 31 \tab 4 \tab 1 \tab 2023 \cr
   Total \tab 1887 \tab 44 \tab 56 \tab 31 \tab 4 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\keyword{variable}
