% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{newowned}
\alias{newowned}
\title{Was the home new or previously owned?}
\source{
General Social Survey https://gss.norc.org
}
\description{
newowned
}
\details{
Question Was the home new or previously owned?
}
\section{Values}{

\itemize{
\item \verb{[1]} new
\item \verb{[2]} previously owned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1815/vshow}{GSS Data Explorer page for \code{newowned}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab new \tab previously owned \tab no answer \tab Total \cr
   1996 \tab 2451 \tab 134 \tab 318 \tab 1 \tab 2904 \cr
   Total \tab 2451 \tab 134 \tab 318 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
