% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_268.R
\name{news12}
\alias{news12}
\title{Past 12 months r web to learn news}
\source{
General Social Survey https://gss.norc.org
}
\description{
news12
}
\details{
Question In the past 12 months, have you used the Web to find out about or discuss?
K. News headlines?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2438/vshow}{GSS Data Explorer page for \code{news12}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 212 \tab 68 \tab 84 \tab 2817 \cr
   2002 \tab 2411 \tab 251 \tab 58 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 463 \tab 126 \tab 129 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{enviro12}},
\code{\link{fe12}},
\code{\link{foraff12}},
\code{\link{gun12}},
\code{\link{moral12}},
\code{\link{polagree}},
\code{\link{polcam12}},
\code{\link{racrel12}},
\code{\link{taxes12}}
}
\concept{Information Society}
\keyword{variable}
