% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{news30}
\alias{news30}
\title{R use news site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
news30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
E. News and current events?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} 1-2 times
\item \verb{[3]} 3-5 times
\item \verb{[4]} more than 5 times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2398/vshow}{GSS Data Explorer page for \code{news30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 153 \tab 186 \tab 120 \tab 209 \tab 74 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 221 \tab 211 \tab 158 \tab 399 \tab 94 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 179 \tab 146 \tab 108 \tab 245 \tab 12 \tab - \tab 2812 \cr
   Total \tab 5878 \tab 553 \tab 543 \tab 386 \tab 853 \tab 180 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{bookmark}},
\code{\link{catdrcty}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{fin30}},
\code{\link{govt30}},
\code{\link{hyperlnk}},
\code{\link{schl30}},
\code{\link{work30}}
}
\concept{Information Society}
\keyword{variable}
