% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{newsfrom}
\alias{newsfrom}
\title{Main source of information about events in the news}
\source{
General Social Survey https://gss.norc.org
}
\description{
newsfrom
}
\details{
Question We are interested in how people get information about events in the news. Where do you get most  of your information about current news events " newspapers, magazines, the Internet, books or other printed materials, TV, radio, government agencies, family, friends, colleagues, or some other source?
}
\section{Values}{

\itemize{
\item \verb{[1]} newspaper
\item \verb{[2]} magazines
\item \verb{[3]} the internet
\item \verb{[4]} books/other printed material
\item \verb{[5]} tv
\item \verb{[6]} radio
\item \verb{[7]} government agencies
\item \verb{[8]} family
\item \verb{[9]} friends/colleagues
\item \verb{[10]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
   C/-/- \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3314/vshow}{GSS Data Explorer page for \code{newsfrom}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab newspaper \tab magazines \tab the internet \tab books/other printed material \tab tv \tab radio \tab government agencies \tab family \tab friends/colleagues \tab other \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 446 \tab 20 \tab 253 \tab 9 \tab 916 \tab 131 \tab 3 \tab 33 \tab 33 \tab 15 \tab 3 \tab 2 \tab 4510 \cr
   2008 \tab 518 \tab 286 \tab 16 \tab 310 \tab 8 \tab 741 \tab 93 \tab 2 \tab 21 \tab 20 \tab 3 \tab 2 \tab 3 \tab 2023 \cr
   2010 \tab 1761 \tab 45 \tab 1 \tab 62 \tab 3 \tab 144 \tab 19 \tab 1 \tab 1 \tab 7 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 972 \tab 121 \tab 6 \tab 299 \tab 5 \tab 476 \tab 54 \tab 1 \tab 17 \tab 13 \tab 2 \tab 4 \tab 4 \tab 1974 \cr
   2014 \tab 1299 \tab 134 \tab 6 \tab 420 \tab 4 \tab 550 \tab 75 \tab 3 \tab 25 \tab 20 \tab 1 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 110 \tab 7 \tab 601 \tab 4 \tab 532 \tab 65 \tab 1 \tab 29 \tab 25 \tab 12 \tab - \tab 4 \tab 2867 \cr
   2018 \tab 1173 \tab 79 \tab 2 \tab 539 \tab 5 \tab 422 \tab 69 \tab 4 \tab 30 \tab 22 \tab 2 \tab 1 \tab - \tab 2348 \cr
   Total \tab 9846 \tab 1221 \tab 58 \tab 2484 \tab 38 \tab 3781 \tab 506 \tab 15 \tab 156 \tab 140 \tab 35 \tab 11 \tab 13 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
