% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_328.R
\name{nextdoor}
\alias{nextdoor}
\title{How willing r move next door to child's family}
\source{
General Social Survey https://gss.norc.org
}
\description{
nextdoor
}
\details{
Question A. How willing would you be to move next door to (John/Mary)''s family? -- definitely willing, probably willing, probably unwilling, or definitely unwilling?
}
\section{Values}{

\itemize{
\item \code{1} definitely willing
\item \code{2} probably willing
\item \code{3} probably unwilling
\item \code{4} definitely unwilling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2933/vshow}{GSS Data Explorer page for \code{nextdoor}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 494 \tab 651 \tab 139 \tab 63 \tab 33 \tab 13 \tab 2765 \cr
   Total \tab 1372 \tab 494 \tab 651 \tab 139 \tab 63 \tab 33 \tab 13 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{allergic}},
\code{\link{bettrown}},
\code{\link{discipln}},
\code{\link{geneprob}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{raised}},
\code{\link{violtv}}
}
\concept{Mental Health II}
\keyword{variable}
