% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{nextgen}
\alias{nextgen}
\title{Science & tech. give more opportunities to next generation}
\source{
General Social Survey https://gss.norc.org
}
\description{
nextgen
}
\details{
Question I’m going to read to you some statements like those you might find in a newspaper or magazine article. For each statement, please tell me if you strongly agree, agree, disagree, or strongly disagree.
A. Because of science and technology, there will be more opportunities for the next generation.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008, 2022 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3317/vshow}{GSS Data Explorer page for \code{nextgen}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2006 \tab 2646 \tab 750 \tab 931 \tab 129 \tab 19 \tab 33 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 518 \tab 542 \tab 787 \tab 113 \tab 18 \tab 42 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 1353 \tab 222 \tab 397 \tab 42 \tab 9 \tab 17 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 277 \tab 582 \tab 90 \tab 13 \tab 34 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 409 \tab 676 \tab 122 \tab 16 \tab 14 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 511 \tab 749 \tab 94 \tab 17 \tab 15 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 455 \tab 625 \tab 76 \tab 14 \tab 4 \tab 1 \tab - \tab 2348 \cr
   2022 \tab 2309 \tab 522 \tab 564 \tab 126 \tab 18 \tab - \tab - \tab 5 \tab 3544 \cr
   Total \tab 11747 \tab 3688 \tab 5311 \tab 792 \tab 124 \tab 159 \tab 22 \tab 5 \tab 21848 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}

Other GSS Next: 
\code{\link{advfront}},
\code{\link{toofast}}
}
\concept{GSS Next}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
