% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{nextstatus}
\alias{nextstatus}
\title{Gss next follow-on status}
\source{
General Social Survey https://gss.norc.org
}
\description{
nextstatus
}
\details{
Question FLAG INDICATING WHETHER OR NOT THE RESPONDENT PARTICIPATED IN THE GSS NEXT FOLLOW-ON
}
\section{Values}{

\itemize{
\item \verb{[1]} completed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7962/vshow}{GSS Data Explorer page for \code{nextstatus}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab completed \tab Total \cr
   2022 \tab 2309 \tab 1235 \tab 3544 \cr
   Total \tab 2309 \tab 1235 \tab 3544 \cr
}
}

\keyword{variable}
