% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{nihilism}
\alias{nihilism}
\title{Life serves no purpose}
\source{
General Social Survey https://gss.norc.org
}
\description{
nihilism
}
\details{
Question Do you agree or disagree with the following. . .
D. In my opinion, life does not serve any purpose.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4660/vshow}{GSS Data Explorer page for \code{nihilism}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 14 \tab 29 \tab 57 \tab 423 \tab 718 \tab 41 \tab 77 \tab 1517 \cr
   1998 \tab 1548 \tab 9 \tab 29 \tab 75 \tab 424 \tab 638 \tab 38 \tab 71 \tab 2832 \cr
   2008 \tab 658 \tab 12 \tab 27 \tab 66 \tab 490 \tab 755 \tab 14 \tab 1 \tab 2023 \cr
   2018 \tab 1173 \tab 12 \tab 22 \tab 67 \tab 396 \tab 661 \tab 8 \tab 9 \tab 2348 \cr
   Total \tab 3537 \tab 47 \tab 107 \tab 265 \tab 1733 \tab 2772 \tab 101 \tab 158 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
