% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{nmbrkids}
\alias{nmbrkids}
\title{Number of children}
\source{
General Social Survey https://gss.norc.org
}
\description{
nmbrkids
}
\details{
Question How many children have you ever had? Please include any biological, adoptive, foster or step-children.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7788/vshow}{GSS Data Explorer page for \code{nmbrkids}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 14 \tab 2 \tab 3 \tab 31 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2386 \tab 556 \tab 2 \tab 1 \tab 1 \tab 291 \tab 160 \tab 1 \tab 70 \tab 37 \tab 13 \tab 9 \tab 3 \tab 1 \tab 4 \tab 9 \tab 3544 \cr
   Total \tab 2386 \tab 556 \tab 2 \tab 1 \tab 1 \tab 291 \tab 160 \tab 1 \tab 70 \tab 37 \tab 13 \tab 9 \tab 3 \tab 1 \tab 4 \tab 9 \tab 3544 \cr
}
}

\keyword{variable}
