% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{noathome}
\alias{noathome}
\title{Miss a family occasion or holiday?}
\source{
General Social Survey https://gss.norc.org
}
\description{
noathome
}
\details{
Question In your (present/most recent) job, (have you ever done/did you ever do) any of the following because of your responsibilities to the job?
A. Miss a family occasion or holiday?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1671/vshow}{GSS Data Explorer page for \code{noathome}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 673 \tab 678 \tab 109 \tab 2904 \cr
   Total \tab 1444 \tab 673 \tab 678 \tab 109 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{earnmore}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}
}
\concept{Gender}
\keyword{variable}
