% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_254.R
\name{nobiling}
\alias{nobiling}
\title{Bilingual education should be abolished}
\source{
General Social Survey https://gss.norc.org
}
\description{
nobiling
}
\details{
Question Now please tell us whether you strongly agree, agree, disagree or strongly disagree with each of these statements:
B. Bilingual education programs should be eliminated in American public schools.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2296/vshow}{GSS Data Explorer page for \code{nobiling}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 80 \tab 218 \tab 665 \tab 380 \tab 50 \tab 5 \tab 2817 \cr
   Total \tab 1419 \tab 80 \tab 218 \tab 665 \tab 380 \tab 50 \tab 5 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang1}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{twolang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
