% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{nobuygrn}
\alias{nobuygrn}
\title{How often avoid products}
\source{
General Social Survey https://gss.norc.org
}
\description{
nobuygrn
}
\details{
Question K3. And how often do you avoid buying certain products for environmental reasons?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4815/vshow}{GSS Data Explorer page for \code{nobuygrn}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{peopgrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}

Other Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{peopgrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
