% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{nocondom}
\alias{nocondom}
\title{Had vaginal or anal sex without a condom in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
nocondom
}
\details{
Question Did you, or did you not, have vaginal or anal sex without a condom in the past 12 months?
}
\section{Values}{

\itemize{
\item \verb{[1]} i had vaginal or anal sex without a condom in the past 12 months
\item \verb{[2]} i did not have vaginal or anal sex without a condom in the past 12 months
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8151/vshow}{GSS Data Explorer page for \code{nocondom}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab i had vaginal or anal sex without a condom in the past 12 months \tab i did not have vaginal or anal sex without a condom in the past 12 months \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 956 \tab 874 \tab 147 \tab 1188 \tab 33 \tab 111 \tab 3309 \cr
   Total \tab 0 \tab 956 \tab 874 \tab 147 \tab 1188 \tab 33 \tab 111 \tab 3309 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{nodocidu}},
\code{\link{sexornt1}},
\code{\link{sexpaid}}
}
\concept{High Risk Behaviors}
\keyword{variable}
