% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{nodocidu}
\alias{nodocidu}
\title{Ever injected any drugs other than those prescribed for you}
\source{
General Social Survey https://gss.norc.org
}
\description{
nodocidu
}
\details{
Question Have you ever in your life injected any drugs other than those prescribed for you? Do not include anything you took under a doctor’s orders.
As of 2024, NODOCIDU has been slightly adjusted to help protect confidentiality. Please see the 2024 codebook for more information.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8200/vshow}{GSS Data Explorer page for \code{nodocidu}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 65 \tab 1950 \tab 56 \tab 1188 \tab 6 \tab 44 \tab 3309 \cr
   Total \tab 0 \tab 65 \tab 1950 \tab 56 \tab 1188 \tab 6 \tab 44 \tab 3309 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{nocondom}},
\code{\link{sexornt1}},
\code{\link{sexpaid}}

Other Health: 
\code{\link{sexpaid}}
}
\concept{Health}
\concept{High Risk Behaviors}
\keyword{variable}
