% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_184.R
\name{noemote}
\alias{noemote}
\title{I keep my emotions to myself.}
\source{
General Social Survey https://gss.norc.org
}
\description{
noemote
}
\details{
Question The following are some statements that people have made. For each one, please tell me if you strongly agree, agree, disagree, or strongly disagree with the statement. (NOTE: INTERVIEWERS SHOULD CODE "neither agree or disagree/it depends" ONLY WHEN VOLUNTEERED.)
G. I keep my emotions to myself.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree or disagree-it depends
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1731/vshow}{GSS Data Explorer page for \code{noemote}}.
}

\seealso{
Other Emotions: 
\code{\link{angry}},
\code{\link{badbrks}},
\code{\link{embarrss}},
\code{\link{litcntrl}},
\code{\link{mostluck}},
\code{\link{noplan}},
\code{\link{proud}},
\code{\link{showangr}},
\code{\link{showfeel}}

Other Feelings: 
\code{\link{angry}},
\code{\link{badbrks}},
\code{\link{embarrss}},
\code{\link{litcntrl}},
\code{\link{mostluck}},
\code{\link{noplan}},
\code{\link{proud}},
\code{\link{showangr}},
\code{\link{showfeel}}
}
\concept{Emotions}
\concept{Feelings}
\keyword{variable}
