% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_236.R
\name{nogo1}
\alias{nogo1}
\title{Performance or exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
nogo1
}
\details{
Question Was it a performance or an exhibit?
}
\section{Values}{

\itemize{
\item \verb{[1]} performance
\item \verb{[2]} exhibit
\item \verb{[3]} both performance and exhibit
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6248/vshow}{GSS Data Explorer page for \code{nogo1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab performance \tab exhibit \tab both performance and exhibit \tab Total \cr
   2016 \tab 2251 \tab 429 \tab 126 \tab 61 \tab 2867 \cr
   Total \tab 2251 \tab 429 \tab 126 \tab 61 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artwhy3}},
\code{\link{artwhy4}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo}},
\code{\link{prfmcost}}
}
\concept{Arts and Culture}
\keyword{variable}
