% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{nohmewrk}
\alias{nohmewrk}
\title{Unable to do the work around the house?}
\source{
General Social Survey https://gss.norc.org
}
\description{
nohmewrk
}
\details{
Question In your (present/most recent) job, (have you ever done/did you ever do) any of the following because of your responsibilities to the job?
C. Been unable to do the work you usually do around the house?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1673/vshow}{GSS Data Explorer page for \code{nohmewrk}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 632 \tab 719 \tab 109 \tab 2904 \cr
   Total \tab 1444 \tab 632 \tab 719 \tab 109 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{chldben}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nonurse}},
\code{\link{parleave}},
\code{\link{womenben}}
}
\concept{Gender}
\keyword{variable}
