% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{nokids}
\alias{nokids}
\title{Desirability of no-kid family}
\source{
General Social Survey https://gss.norc.org
}
\description{
nokids
}
\details{
Question In general, what do you feel about each of these family sizes?
A. A family with no children
}
\section{Values}{

\itemize{
\item \code{1} very desirable
\item \code{2} desirable
\item \code{3} neither
\item \code{4} undesirable
\item \code{5} very undesirable
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4375/vshow}{GSS Data Explorer page for \code{nokids}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very desirable \tab desirable \tab neither \tab undesirable \tab very undesirable \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 53 \tab 167 \tab 473 \tab 366 \tab 162 \tab 114 \tab 79 \tab 1481 \cr
   Total \tab 67 \tab 53 \tab 167 \tab 473 \tab 366 \tab 162 \tab 114 \tab 79 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{mardiv}},
\code{\link{marhomo}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{threkids}},
\code{\link{twokids}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
