% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{nomeat}
\alias{nomeat}
\title{Refuse to eat meat for envir reasons}
\source{
General Social Survey https://gss.norc.org
}
\description{
nomeat
}
\details{
Question C. And how often do you refuse to eat meat for moral or environmental reasons?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} never
\item \verb{[5]} no car, dont drive
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4771/vshow}{GSS Data Explorer page for \code{nomeat}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab never \tab no answer \tab Total \cr
   1993 \tab 49 \tab 35 \tab 105 \tab 322 \tab 1056 \tab 39 \tab 1606 \cr
   1994 \tab 1606 \tab 41 \tab 99 \tab 272 \tab 952 \tab 22 \tab 2992 \cr
   Total \tab 1655 \tab 76 \tab 204 \tab 594 \tab 2008 \tab 61 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{recycle}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
