% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{nonbinkids}
\alias{nonbinkids}
\title{First child nonbinary}
\source{
General Social Survey https://gss.norc.org
}
\description{
nonbinkids
}
\details{
Question Are any of your biological children transgender or non-binary?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7959/vshow}{GSS Data Explorer page for \code{nonbinkids}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab Total \cr
   2022 \tab 2308 \tab 18 \tab 1171 \tab 47 \tab 3544 \cr
   Total \tab 2308 \tab 18 \tab 1171 \tab 47 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{femself}},
\code{\link{firstkidsex}},
\code{\link{mascself}}
}
\concept{GSS Next}
\keyword{variable}
