% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{nonuke}
\alias{nonuke}
\title{Possibility eliminating atomic weapons}
\source{
General Social Survey https://gss.norc.org
}
\description{
nonuke
}
\details{
Question I'm going to read you some possible military situations the U.S. might face in the next ten years. (HAND CARD BB)
Some people feel these situations are certain to happen (think of these as point 7 on the scale), others think these situations won't happen at all (think of these as point 1 on the scale). And of course some people have opinions somewhere in between. For each of these possible military situations, please give me your best guess as to how likely it is to happen. For example, an all-out atomic war. Where would you put the likelihood of an all-out atomic war during the next ten years? (Would you say 7--it is certain to happen, 1--it won't happen at all, or something in between?)
(CIRCLE ONE CODE FOR "A" AND REPEAT FOR B-F)
E. Elimination of atomic weapons by both U.S. and Russia?
}
\section{Values}{

\itemize{
\item \code{1} wont happen
\item \code{7} certain to happen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/826/vshow}{GSS Data Explorer page for \code{nonuke}}.
}

\seealso{
Other Core: 
\code{\link{defwrkev}},
\code{\link{guerilla}},
\code{\link{jobtrain}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()},
\code{\link{vetaid}}
}
\concept{Core}
\keyword{variable}
