% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{nonurse}
\alias{nonurse}
\title{Unable to care for a sick child-relative?}
\source{
General Social Survey https://gss.norc.org
}
\description{
nonurse
}
\details{
Question In your (present/most recent) job, (have you ever done/did you ever do) any of the following because of your responsibilities to the job?
B. Been unable to care for a sick child or relative?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1672/vshow}{GSS Data Explorer page for \code{nonurse}}.
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{chldben}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{parleave}},
\code{\link{womenben}}

Other Family: 
\code{\link{allben}},
\code{\link{chldben}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{parleave}},
\code{\link{womenben}}

Other Job: 
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{flextime}},
\code{\link{nohmewrk}},
\code{\link{parleave}}

Other Work: 
\code{\link{flextime}},
\code{\link{nohmewrk}},
\code{\link{parleave}}
}
\concept{Family}
\concept{Gender}
\concept{Job}
\concept{Work}
\keyword{variable}
