% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_180.R
\name{noonehrt}
\alias{noonehrt}
\title{No one hurt from traditional family}
\source{
General Social Survey https://gss.norc.org
}
\description{
noonehrt
}
\details{
Question Who is hurt by this arrangement: men, women, children, everyone, or no one? PROBE:  Anyone else? UNTIL R SAYS NO.
E. No one
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1687/vshow}{GSS Data Explorer page for \code{noonehrt}}.
}

\seealso{
Other Gender: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{fekids1}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{womenhrt}}

Other Family: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{womenhrt}}

Other Household: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{fekids1}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{womenhrt}}
}
\concept{Family}
\concept{Gender}
\concept{Household}
\keyword{variable}
