% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{norelgsp16}
\alias{norelgsp16}
\title{R no religion age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
norelgsp16
}
\details{
Question Were  you raised to be atheist, agnostic, something else, or nothing in particular?
}
\section{Values}{

\itemize{
\item \verb{[1]} atheist
\item \verb{[2]} agnostic
\item \verb{[3]} something else
\item \verb{[4]} nothing in particular
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8240/vshow}{GSS Data Explorer page for \code{norelgsp16}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab atheist \tab agnostic \tab something else \tab nothing in particular \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 15 \tab 19 \tab 18 \tab 375 \tab 4 \tab 2875 \tab 1 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 15 \tab 19 \tab 18 \tab 375 \tab 4 \tab 2875 \tab 1 \tab 2 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conlabory}},
\code{\link{marsamey}},
\code{\link{pawrkgrw}},
\code{\link{racdif5}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
