% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_439.R
\name{nosmoke}
\alias{nosmoke}
\title{Gov should prohibit smoking in pub areas}
\source{
General Social Survey https://gss.norc.org
}
\description{
nosmoke
}
\details{
Question Do you think that...
B. Smoking in public places should be prohibited by law.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4037/vshow}{GSS Data Explorer page for \code{nosmoke}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab 5 \tab no answer \tab Total \cr
   1985 \tab 857 \tab 134 \tab 170 \tab 150 \tab 167 \tab 47 \tab 9 \tab 1534 \cr
   Total \tab 857 \tab 134 \tab 170 \tab 150 \tab 167 \tab 47 \tab 9 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{beltup}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidout}},
\code{\link{kidskips}},
\code{\link{kidxfilm}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
