% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{notam}
\alias{notam}
\title{If not american, how much change}
\source{
General Social Survey https://gss.norc.org
}
\description{
notam
}
\details{
Question If you were no longer an AMERICAN, but were a citizen of another country living outside the United States, how much would it change what you Do in a typical week?  Would what you do in a typical week be:
}
\section{Values}{

\itemize{
\item \verb{[1]} completely different
\item \verb{[2]} quite different
\item \verb{[3]} somewhat different
\item \verb{[4]} just slightly different
\item \verb{[5]} not at all different
\item \verb{[7]} citizen of another country
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1788/vshow}{GSS Data Explorer page for \code{notam}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab completely different \tab quite different \tab somewhat different \tab just slightly different \tab not at all different \tab citizen of another country \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2199 \tab 191 \tab 123 \tab 148 \tab 68 \tab 81 \tab 13 \tab 76 \tab 5 \tab 2904 \cr
   Total \tab 2199 \tab 191 \tab 123 \tab 148 \tab 68 \tab 81 \tab 13 \tab 76 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
