% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{notreat}
\alias{notreat}
\title{Refuse treatment except pain relief}
\source{
General Social Survey https://gss.norc.org
}
\description{
notreat
}
\details{
Question How much do you agree of disagree with the following options for this person?  This person should...
B. Refuse further treatment except for pain relief right now.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2256/vshow}{GSS Data Explorer page for \code{notreat}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 444 \tab 523 \tab 165 \tab 151 \tab 32 \tab 33 \tab 39 \tab 2832 \cr
   Total \tab 1445 \tab 444 \tab 523 \tab 165 \tab 151 \tab 32 \tab 33 \tab 39 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{painmed}},
\code{\link{rightref}},
\code{\link{termvig}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\keyword{variable}
