% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{nounemp1}
\alias{nounemp1}
\title{R would accept job requiring new skill to avoid unemployment}
\source{
General Social Survey https://gss.norc.org
}
\description{
nounemp1
}
\details{
Question To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing…
A. To accept a job that requires new skills
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4620/vshow}{GSS Data Explorer page for \code{nounemp1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 459 \tab 416 \tab 66 \tab 53 \tab 17 \tab 5 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 381 \tab 429 \tab 59 \tab 45 \tab 14 \tab 5 \tab 7 \tab 2867 \cr
   Total \tab 5419 \tab 840 \tab 845 \tab 125 \tab 98 \tab 31 \tab 10 \tab 9 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
