% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{nounemp2}
\alias{nounemp2}
\title{R would accept position with lower pay to avoid unemployment}
\source{
General Social Survey https://gss.norc.org
}
\description{
nounemp2
}
\details{
Question To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing…
B. To accept a position with lower pay
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4621/vshow}{GSS Data Explorer page for \code{nounemp2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 183 \tab 423 \tab 129 \tab 191 \tab 84 \tab 6 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 149 \tab 362 \tab 128 \tab 209 \tab 82 \tab 3 \tab 7 \tab 2867 \cr
   Total \tab 5419 \tab 332 \tab 785 \tab 257 \tab 400 \tab 166 \tab 9 \tab 9 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
