% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{nounemp4}
\alias{nounemp4}
\title{R would accept travel longer to work to avoid unemployment}
\source{
General Social Survey https://gss.norc.org
}
\description{
nounemp4
}
\details{
Question To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing…
D. To travel longer to get to work
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4623/vshow}{GSS Data Explorer page for \code{nounemp4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 158 \tab 438 \tab 101 \tab 231 \tab 86 \tab 2 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 131 \tab 422 \tab 127 \tab 185 \tab 63 \tab 4 \tab 8 \tab 2867 \cr
   Total \tab 5419 \tab 289 \tab 860 \tab 228 \tab 416 \tab 149 \tab 6 \tab 10 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{nounemp3}},
\code{\link{otherwrk}},
\code{\link{skltrain}},
\code{\link{spwrkgvt}}
}
\concept{Work Orientation}
\keyword{variable}
