% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{nounemp6}
\alias{nounemp6}
\title{I would move outside the usa for work}
\source{
General Social Survey https://gss.norc.org
}
\description{
nounemp6
}
\details{
Question (HAND CARD D17)
(To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing...)
To move to a different country.
(Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6336/vshow}{GSS Data Explorer page for \code{nounemp6}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 43 \tab 120 \tab 131 \tab 307 \tab 328 \tab 4 \tab 7 \tab 2867 \cr
   Total \tab 1927 \tab 43 \tab 120 \tab 131 \tab 307 \tab 328 \tab 4 \tab 7 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
