% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_673.R
\name{nouseint10}
\alias{nouseint10}
\title{No internet use over past 12 months due to lack of time}
\source{
General Social Survey https://gss.norc.org
}
\description{
nouseint10
}
\details{
Question Which of the following are reasons you have not used the Internet in the last 12 months?… I am too busy to use the Internet.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8206/vshow}{GSS Data Explorer page for \code{nouseint10}}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab yes \tab no \tab can't choose \tab iap \tab Total \cr
   2024 \tab - \tab 14 \tab 41 \tab 9 \tab 3245 \tab 3309 \cr
   Total \tab 0 \tab 14 \tab 41 \tab 9 \tab 3245 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{ftruse}},
\code{\link{intaccprxy}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint6}},
\code{\link{nouseint7}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{prxyuse}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
