% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_673.R
\name{nouseint12}
\alias{nouseint12}
\title{No internet use over past 12 months due to other reasons}
\source{
General Social Survey https://gss.norc.org
}
\description{
nouseint12
}
\details{
Question Which of the following are reasons you have not used the Internet in the last 12 months?… Other reasons
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8208/vshow}{GSS Data Explorer page for \code{nouseint12}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{ftruse}},
\code{\link{intaccprxy}},
\code{\link{inthome}},
\code{\link{intprfssnl}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{prxyuse}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
