% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{nouseint3}
\alias{nouseint3}
\title{No internet use over past 12 months due to cost}
\source{
General Social Survey https://gss.norc.org
}
\description{
nouseint3
}
\details{
Question Which of the following are reasons you have not used the Internet in the last 12 months?… The Internet is too expensive.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8210/vshow}{GSS Data Explorer page for \code{nouseint3}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab yes \tab no \tab can't choose \tab iap \tab no answer \tab Total \cr
   2024 \tab - \tab 21 \tab 32 \tab 10 \tab 3245 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 21 \tab 32 \tab 10 \tab 3245 \tab 1 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intrnetuse}},
\code{\link{nouseint1}},
\code{\link{nouseint2}},
\code{\link{nouseint4}},
\code{\link{nouseint5}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
