% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{nukeacc}
\alias{nukeacc}
\title{Likelihood of nuclear meltdown in 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukeacc
}
\details{
Question Within the next five years, how likely is it that an accident at a nuclear power station will cause long-term environmental damage across many countries?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} likely
\item \code{3} unlikely
\item \code{4} very unlikely
\item \code{5} cant choose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4790/vshow}{GSS Data Explorer page for \code{nukeacc}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{econgrn}},
\code{\link{excldimm}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{ldcgrn}}

Other Environment: 
\code{\link{econgrn}},
\code{\link{ldcgrn}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
