% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{nukeelec}
\alias{nukeelec}
\title{View about nuclear energy}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukeelec
}
\details{
Question Please look at Card E16. When making policy recommendations about nuclear power, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests? FI INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
D. Which statement best describes your own view about increasing the use of nuclear power to generate electricity in the U.S.?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} oppose
\item \verb{[4]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3498/vshow}{GSS Data Explorer page for \code{nukeelec}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1578 \tab 56 \tab 214 \tab 104 \tab 37 \tab 47 \tab 8 \tab 2044 \cr
   Total \tab 1578 \tab 56 \tab 214 \tab 104 \tab 37 \tab 47 \tab 8 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstnk}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
