% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{nukegen}
\alias{nukegen}
\title{Nuke power danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukegen
}
\details{
Question A. In general, do you think that nuclear power stations are…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4756/vshow}{GSS Data Explorer page for \code{nukegen}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 287 \tab 334 \tab 540 \tab 206 \tab 36 \tab 65 \tab 89 \tab 1606 \cr
   1994 \tab 1606 \tab 263 \tab 300 \tab 489 \tab 174 \tab 25 \tab 65 \tab 70 \tab 2992 \cr
   2010 \tab 614 \tab 277 \tab 336 \tab 442 \tab 210 \tab 66 \tab 95 \tab 4 \tab 2044 \cr
   2016 \tab 1956 \tab 246 \tab 234 \tab 298 \tab 102 \tab 15 \tab 14 \tab 2 \tab 2867 \cr
   2018 \tab 1563 \tab 218 \tab 213 \tab 257 \tab 64 \tab 20 \tab 12 \tab 1 \tab 2348 \cr
   Total \tab 5788 \tab 1291 \tab 1417 \tab 2026 \tab 756 \tab 162 \tab 251 \tab 166 \tab 11857 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}}
}
\concept{ISSP Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
