% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{nukegen1}
\alias{nukegen1}
\title{Nuke power danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukegen1
}
\details{
Question And do you think that nuclear power stations are…
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7494/vshow}{GSS Data Explorer page for \code{nukegen1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 309 \tab 386 \tab 631 \tab 328 \tab 89 \tab 88 \tab 188 \tab 10 \tab 4032 \cr
   Total \tab 2003 \tab 309 \tab 386 \tab 631 \tab 328 \tab 89 \tab 88 \tab 188 \tab 10 \tab 4032 \cr
}
}

\keyword{variable}
