% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{numbwfrd}
\alias{numbwfrd}
\title{Good friends who are white-black}
\source{
General Social Survey https://gss.norc.org
}
\description{
numbwfrd
}
\details{
Question How many of your good friends are White/Black?
}
\section{Values}{

\itemize{
\item \verb{[94]} no numuber given
\item \verb{[95]} many
\item \verb{[96]} 96+
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1135/vshow}{GSS Data Explorer page for \code{numbwfrd}}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 15 \tab 2 \tab 20 \tab 25 \tab 27 \tab 3 \tab 30 \tab 35 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 1976 \tab 580 \tab 117 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 62 \tab 3 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 12 \tab 10 \tab 4 \tab 4 \tab 1 \tab 1 \tab 20 \tab - \tab 2832 \cr
   Total \tab 1976 \tab 580 \tab 117 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 62 \tab 3 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 12 \tab 10 \tab 4 \tab 4 \tab 1 \tab 1 \tab 20 \tab 0 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{madenkid}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{othsch}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcand}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}

Other Friends: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{bwfriend}},
\code{\link{clsclsex}},
\code{\link{cowrkfrd}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdenjoy}},
\code{\link{frdfun}},
\code{\link{frdhelps}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdknows}},
\code{\link{frdresp}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{friends}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndknow}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{makefrnd}},
\code{\link{needyfrd}},
\code{\link{neifrd}},
\code{\link{numfrend}},
\code{\link{othfrd}},
\code{\link{rightpeo}}

Other Race: 
\code{\link{admirblk}},
\code{\link{allowmar}},
\code{\link{asngovt}},
\code{\link{blkgains}},
\code{\link{blkgovt}},
\code{\link{blkinflu}},
\code{\link{blksimp}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{bwfriend}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{color}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{dsndntpay}},
\code{\link{ethchng}},
\code{\link{ethhist}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdmar}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{hseth}},
\code{\link{hspgovt}},
\code{\link{hswht}},
\code{\link{leastcom}},
\code{\link{meltpot}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{oprace}},
\code{\link{owneth}},
\code{\link{perla}},
\code{\link{perla_self}},
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs16}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs7}},
\code{\link{raceacs9}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmeet}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpromo}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{ractrust}},
\code{\link{racwork}},
\code{\link{ratetone}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{trtmyrac}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wrkwayup}}
}
\concept{Friends}
\concept{Race}
\concept{Religion}
\keyword{variable}
