% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_430.R
\name{numclerg}
\alias{numclerg}
\title{With how many different leaders has this happened to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
numclerg
}
\details{
Question With how may different leaders has this happened to you?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3938/vshow}{GSS Data Explorer page for \code{numclerg}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 5 \tab Total \cr
   2008 \tab 1999 \tab 17 \tab 5 \tab 1 \tab 1 \tab 2023 \cr
   Total \tab 1999 \tab 17 \tab 5 \tab 1 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
