% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{numdays}
\alias{numdays}
\title{Number of days attended religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
numdays
}
\details{
Question On what day or days do you attend religious services during the last seven days? (PROBE, ASK UNTIL "NO" IS GIVEN: "Did you attend religious services on any other days during the last seven days?")
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5310/vshow}{GSS Data Explorer page for \code{numdays}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab no answer \tab Total \cr
   1996 \tab 2608 \tab 1 \tab 185 \tab 42 \tab 16 \tab 3 \tab 1 \tab 2 \tab 4 \tab 42 \tab 2904 \cr
   1998 \tab 2329 \tab 2 \tab 336 \tab 105 \tab 29 \tab 12 \tab 3 \tab 1 \tab 8 \tab 7 \tab 2832 \cr
   Total \tab 4937 \tab 3 \tab 521 \tab 147 \tab 45 \tab 15 \tab 4 \tab 3 \tab 12 \tab 49 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attrelig}},
\code{\link{eatout}},
\code{\link{godoc}},
\code{\link{monday}},
\code{\link{seefilm}},
\code{\link{sunday}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
