% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{numemps}
\alias{numemps}
\title{Number of employee for the self-employed}
\source{
General Social Survey https://gss.norc.org
}
\description{
numemps
}
\details{
Question Are/were you self-employed or do/did you work for someone else?
(If "Self-employed") Do/did you have any employees?
(If "Yes") How many employees do/did you have?
}
\section{Values}{

\itemize{
\item \verb{[0]} 0 employees
\item \verb{[1]} 1-9 employees
\item \verb{[2]} 10+ employees
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3014/vshow}{GSS Data Explorer page for \code{numemps}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab 0 employees \tab 1-9 employees \tab 10+ employees \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2004 \tab - \tab 209 \tab 93 \tab 38 \tab 2467 \tab 5 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 317 \tab 155 \tab 37 \tab 3999 \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 143 \tab 64 \tab 20 \tab 1796 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 153 \tab 62 \tab 19 \tab 1809 \tab - \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab 128 \tab 48 \tab 18 \tab 1779 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 201 \tab 66 \tab 27 \tab 2242 \tab 1 \tab 1 \tab - \tab 2538 \cr
   2016 \tab - \tab 201 \tab 70 \tab 23 \tab 2570 \tab 2 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 149 \tab 65 \tab 20 \tab 2114 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 293 \tab 93 \tab 35 \tab 3601 \tab 3 \tab - \tab 7 \tab 4032 \cr
   2022 \tab - \tab 274 \tab 84 \tab 31 \tab 3148 \tab 2 \tab - \tab 5 \tab 3544 \cr
   Total \tab 0 \tab 2068 \tab 800 \tab 268 \tab 25525 \tab 16 \tab 3 \tab 12 \tab 28692 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Business}
\concept{Employment}
\concept{Workplace Conflict}
\keyword{variable}
