% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{numemps5}
\alias{numemps5}
\title{How many people r employs 5 yrs after}
\source{
General Social Survey https://gss.norc.org
}
\description{
numemps5
}
\details{
Question How many people do you anticipate employing five years from now?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3823/vshow}{GSS Data Explorer page for \code{numemps5}}.
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}}

Other Business: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}}

Other Employment: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{paidemps}},
\code{\link{whybiz}}

Other Job: 
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}}
}
\concept{Business}
\concept{Employment}
\concept{Job}
\concept{Self Employment}
\keyword{variable}
