% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_349.R
\name{numfarm}
\alias{numfarm}
\title{How many r belongs to farm organizations}
\source{
General Social Survey https://gss.norc.org
}
\description{
numfarm
}
\details{
Question L. How many farm organizations do you belong to?
}
\section{Values}{

\itemize{
\item \verb{[7]} seven or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3138/vshow}{GSS Data Explorer page for \code{numfarm}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab no answer \tab Total \cr
   2004 \tab 2763 \tab 41 \tab 2 \tab 1 \tab 5 \tab 2812 \cr
   Total \tab 2763 \tab 41 \tab 2 \tab 1 \tab 5 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{numchurh}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numinfrm}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numprof}},
\code{\link{numschl}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
