% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{numfrend}
\alias{numfrend}
\title{About how many good friends do you have}
\source{
General Social Survey https://gss.norc.org
}
\description{
numfrend
}
\details{
Question B. About how many good friends do you have?
}
\section{Values}{

\itemize{
\item \code{94} no number given
\item \code{95} many
\item \code{96} 96+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1134/vshow}{GSS Data Explorer page for \code{numfrend}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 12 \tab 15 \tab 2 \tab 20 \tab 23 \tab 24 \tab 25 \tab 3 \tab 30 \tab 4 \tab 40 \tab 41 \tab 45 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 75 \tab 8 \tab 9 \tab no number given \tab many \tab 96+ \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1976 \tab 51 \tab 79 \tab 35 \tab 13 \tab 136 \tab 27 \tab 1 \tab 1 \tab 4 \tab 132 \tab 4 \tab 119 \tab 3 \tab 1 \tab 1 \tab 95 \tab 5 \tab 74 \tab 1 \tab 10 \tab 1 \tab 20 \tab 6 \tab 6 \tab 8 \tab 7 \tab 9 \tab 7 \tab 2832 \cr
   Total \tab 1976 \tab 51 \tab 79 \tab 35 \tab 13 \tab 136 \tab 27 \tab 1 \tab 1 \tab 4 \tab 132 \tab 4 \tab 119 \tab 3 \tab 1 \tab 1 \tab 95 \tab 5 \tab 74 \tab 1 \tab 10 \tab 1 \tab 20 \tab 6 \tab 6 \tab 8 \tab 7 \tab 9 \tab 7 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numbwfrd}}
}
\concept{Religion}
\keyword{variable}
