% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{numgiven}
\alias{numgiven}
\title{Number of persons mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
numgiven
}
\details{
Question From time to time, most people discuss important matters with other people. Looking back over the last six months - who are the people with whom you discussed matters important to you? Just tell me their first names or initials. If LESS THAN 5 NAMES MENTIONED, PROBE, Anyone else?  ONLY RECORD FIRST 5 NAMES.
LIST ALL NAMES IN ORDER ACROSS THE TOP OF THE MATRIX (SEE 2 PAGES AHEAD). THEN WRITE NAMES 2-5 DOWN THE SIDE OF THE MATRIX.
A. INTERVIEWER CHECK: HOW MANY NAMES WERE MENTIONED?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/848/vshow}{GSS Data Explorer page for \code{numgiven}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no answer \tab Total \cr
   1985 \tab - \tab 136 \tab 228 \tab 235 \tab 321 \tab 233 \tab 294 \tab 84 \tab 3 \tab 1534 \cr
   1987 \tab - \tab 98 \tab 287 \tab 350 \tab 849 \tab 112 \tab 53 \tab 51 \tab 19 \tab 1819 \cr
   2004 \tab 1340 \tab 356 \tab 281 \tab 263 \tab 232 \tab 128 \tab 96 \tab 70 \tab 46 \tab 2812 \cr
   Total \tab 1340 \tab 590 \tab 796 \tab 848 \tab 1402 \tab 473 \tab 443 \tab 205 \tab 68 \tab 6165 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{eqclose}}
}
\concept{Social Networks}
\keyword{variable}
