% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{numhome}
\alias{numhome}
\title{How many computers at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
numhome
}
\details{
Question C. How many computers in working condition do you have in your home?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2551/vshow}{GSS Data Explorer page for \code{numhome}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2418 \tab 261 \tab 76 \tab 27 \tab 7 \tab 2 \tab 1 \tab 1 \tab 24 \tab 2817 \cr
   Total \tab 2418 \tab 261 \tab 76 \tab 27 \tab 7 \tab 2 \tab 1 \tab 1 \tab 24 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
