% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{nummen}
\alias{nummen}
\title{Number of male sex partner's since 18}
\source{
General Social Survey https://gss.norc.org
}
\description{
nummen
}
\details{
Question Now thinking about the time since your 18th birthday (including the past 12 months) how many male partners have you had sex with?
}
\section{Values}{

\itemize{
\item \verb{[989]} 989 or more
\item \verb{[990]} dash or slash
\item \verb{[991]} 1 or more, number unknown
\item \verb{[992]} 'x'
\item \verb{[993]} garbled text
\item \verb{[994]} 'several'
\item \verb{[995]} 'many,' 'lots'
\item \verb{[996]} 'n.a.'
\item \verb{[997]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5061/vshow}{GSS Data Explorer page for \code{nummen}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 14 \tab 15 \tab 16 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 300 \tab 35 \tab 4 \tab 40 \tab 5 \tab 55 \tab 6 \tab 7 \tab 8 \tab 9 \tab dash or slash \tab garbled text \tab 'many,' 'lots' \tab don't know \tab no answer \tab 19 \tab 23 \tab 45 \tab 50 \tab refused \tab 100 \tab 13 \tab 22 \tab 26 \tab 375 \tab 90 \tab 'x' \tab 'several' \tab 'n.a.' \tab 17 \tab 18 \tab 200 \tab 21 \tab 336 \tab 1 or more, number unknown \tab 150 \tab 24 \tab 34 \tab 350 \tab 36 \tab 60 \tab 70 \tab 27 \tab 33 \tab 65 \tab 41 \tab 43 \tab 500 \tab 51 \tab 75 \tab 120 \tab 122 \tab 28 \tab 32 \tab 400 \tab 37 \tab 42 \tab 48 \tab 103 \tab 110 \tab 39 \tab 989 or more \tab 175 \tab 52 \tab 54 \tab 600 \tab 99 \tab 62 \tab 80 \tab 700 \tab 85 \tab 84 \tab 83 \tab 125 \tab 365 \tab 38 \tab 397 \tab 88 \tab 56 \tab 57 \tab 123 \tab 140 \tab 250 \tab 29 \tab 47 \tab Total \cr
   1989 \tab 136 \tab 574 \tab 324 \tab 22 \tab 1 \tab 4 \tab 2 \tab 11 \tab 1 \tab 99 \tab 9 \tab 4 \tab 75 \tab 5 \tab 1 \tab 1 \tab 50 \tab 1 \tab 34 \tab 1 \tab 27 \tab 11 \tab 10 \tab 3 \tab 6 \tab 1 \tab 2 \tab 9 \tab 113 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 199 \tab 494 \tab 254 \tab 15 \tab - \tab 10 \tab 1 \tab 5 \tab - \tab 81 \tab 7 \tab 6 \tab 67 \tab - \tab - \tab - \tab 43 \tab 2 \tab 32 \tab - \tab 21 \tab 6 \tab 9 \tab 3 \tab 1 \tab - \tab 2 \tab 5 \tab 100 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 221 \tab 496 \tab 280 \tab 27 \tab 2 \tab 5 \tab - \tab 10 \tab - \tab 117 \tab 11 \tab 2 \tab 84 \tab 4 \tab - \tab 2 \tab 40 \tab 3 \tab 54 \tab - \tab 20 \tab 13 \tab 9 \tab 4 \tab 2 \tab - \tab - \tab 5 \tab 87 \tab - \tab 1 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 114 \tab 615 \tab 298 \tab 33 \tab 7 \tab 9 \tab - \tab 10 \tab 2 \tab 123 \tab 6 \tab 4 \tab 85 \tab 2 \tab - \tab - \tab 56 \tab - \tab 43 \tab - \tab 35 \tab 20 \tab 11 \tab 5 \tab 1 \tab - \tab 2 \tab 9 \tab 85 \tab - \tab - \tab 2 \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 201 \tab 1069 \tab 536 \tab 72 \tab 2 \tab 21 \tab 4 \tab 26 \tab 1 \tab 212 \tab 24 \tab 9 \tab 166 \tab 4 \tab 1 \tab 2 \tab 102 \tab 3 \tab 93 \tab - \tab 55 \tab 31 \tab 24 \tab 7 \tab 12 \tab 2 \tab - \tab 14 \tab 217 \tab - \tab 1 \tab - \tab 11 \tab 30 \tab 4 \tab 5 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 7 \tab - \tab 3 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 258 \tab 1033 \tab 431 \tab 85 \tab 1 \tab 16 \tab 2 \tab 23 \tab 2 \tab 187 \tab 30 \tab 7 \tab 151 \tab 14 \tab - \tab 1 \tab 106 \tab 2 \tab 82 \tab - \tab 49 \tab 25 \tab 28 \tab 14 \tab 4 \tab 3 \tab 11 \tab 11 \tab 215 \tab 1 \tab 1 \tab 2 \tab 3 \tab 25 \tab 3 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab 52 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 383 \tab 990 \tab 439 \tab 59 \tab 1 \tab 9 \tab 2 \tab 23 \tab 1 \tab 176 \tab 21 \tab 4 \tab 149 \tab 10 \tab 2 \tab - \tab 96 \tab - \tab 114 \tab - \tab 62 \tab 39 \tab 19 \tab 9 \tab 1 \tab - \tab 4 \tab 14 \tab 162 \tab - \tab - \tab - \tab 2 \tab 9 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 10 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 417 \tab 906 \tab 386 \tab 71 \tab 4 \tab 17 \tab 3 \tab 19 \tab - \tab 171 \tab 25 \tab 7 \tab 132 \tab 7 \tab - \tab 1 \tab 111 \tab 3 \tab 102 \tab - \tab 60 \tab 24 \tab 23 \tab 9 \tab 4 \tab - \tab 4 \tab 12 \tab 228 \tab - \tab - \tab 1 \tab 7 \tab 27 \tab 4 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 489 \tab 982 \tab 332 \tab 62 \tab 3 \tab 13 \tab 1 \tab 28 \tab 1 \tab 156 \tab 15 \tab 13 \tab 138 \tab 9 \tab - \tab 1 \tab 84 \tab 2 \tab 128 \tab - \tab 52 \tab 33 \tab 19 \tab 8 \tab - \tab - \tab 2 \tab 31 \tab 131 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 584 \tab 974 \tab 359 \tab 65 \tab 3 \tab 18 \tab 5 \tab 18 \tab - \tab 153 \tab 14 \tab 10 \tab 127 \tab 9 \tab - \tab 1 \tab 81 \tab 5 \tab 84 \tab - \tab 64 \tab 34 \tab 19 \tab 15 \tab - \tab - \tab - \tab 30 \tab 90 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 9 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 15 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2096 \tab 1028 \tab 379 \tab 64 \tab 2 \tab 11 \tab 2 \tab 25 \tab 3 \tab 174 \tab 31 \tab 10 \tab 152 \tab 6 \tab 1 \tab 2 \tab 112 \tab 2 \tab 107 \tab - \tab 56 \tab 36 \tab 24 \tab 7 \tab - \tab - \tab 4 \tab 18 \tab 100 \tab - \tab 3 \tab 2 \tab 10 \tab 3 \tab 8 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 240 \tab 757 \tab 314 \tab 50 \tab 2 \tab 11 \tab 3 \tab 13 \tab 3 \tab 137 \tab 23 \tab 7 \tab 102 \tab 8 \tab - \tab 4 \tab 68 \tab 4 \tab 71 \tab - \tab 35 \tab 12 \tab 17 \tab 6 \tab - \tab - \tab 2 \tab 30 \tab 60 \tab - \tab - \tab 2 \tab 5 \tab 2 \tab 6 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 202 \tab 809 \tab 247 \tab 58 \tab 5 \tab 9 \tab 2 \tab 24 \tab 2 \tab 121 \tab 18 \tab 7 \tab 114 \tab 5 \tab - \tab 1 \tab 82 \tab 4 \tab 84 \tab - \tab 53 \tab 20 \tab 21 \tab 7 \tab - \tab - \tab - \tab 24 \tab 88 \tab - \tab - \tab - \tab 9 \tab 6 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 238 \tab 715 \tab 253 \tab 46 \tab 3 \tab 16 \tab 1 \tab 19 \tab - \tab 113 \tab 17 \tab 12 \tab 102 \tab 8 \tab 2 \tab 1 \tab 80 \tab 4 \tab 70 \tab - \tab 39 \tab 21 \tab 23 \tab 13 \tab - \tab 1 \tab 4 \tab 30 \tab 88 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 22 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 188 \tab 996 \tab 337 \tab 69 \tab 8 \tab 9 \tab 2 \tab 16 \tab 5 \tab 168 \tab 23 \tab 13 \tab 158 \tab 9 \tab 1 \tab 2 \tab 89 \tab 1 \tab 126 \tab - \tab 50 \tab 32 \tab 36 \tab 7 \tab - \tab - \tab 1 \tab 116 \tab 28 \tab - \tab 1 \tab 1 \tab 9 \tab 1 \tab 5 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1096 \tab 757 \tab 242 \tab 54 \tab 3 \tab 10 \tab 2 \tab 18 \tab 3 \tab 112 \tab 27 \tab 7 \tab 115 \tab 9 \tab - \tab 4 \tab 78 \tab 3 \tab 87 \tab 1 \tab 51 \tab 17 \tab 16 \tab 11 \tab - \tab 1 \tab - \tab 9 \tab 83 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 6 \tab 7 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 942 \tab 607 \tab 200 \tab 50 \tab 3 \tab 10 \tab 3 \tab 23 \tab - \tab 86 \tab 16 \tab 6 \tab 69 \tab 5 \tab - \tab - \tab 76 \tab 2 \tab 74 \tab - \tab 33 \tab 20 \tab 13 \tab 2 \tab - \tab - \tab - \tab 15 \tab 51 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 16 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1705 \tab 965 \tab 309 \tab 98 \tab 2 \tab 30 \tab 4 \tab 30 \tab 6 \tab 137 \tab 39 \tab 12 \tab 111 \tab 23 \tab - \tab - \tab 75 \tab 1 \tab 125 \tab - \tab 49 \tab 36 \tab 38 \tab 14 \tab - \tab - \tab 3 \tab 24 \tab 113 \tab - \tab 1 \tab 2 \tab 13 \tab 1 \tab 8 \tab 7 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 6 \tab 4 \tab 1 \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4032 \cr
   2022 \tab 1898 \tab 654 \tab 193 \tab 42 \tab 3 \tab 10 \tab 1 \tab 14 \tab 4 \tab 89 \tab 21 \tab 11 \tab 90 \tab 8 \tab - \tab 2 \tab 213 \tab 1 \tab 74 \tab - \tab 45 \tab 21 \tab 13 \tab 5 \tab - \tab - \tab - \tab - \tab 105 \tab - \tab 2 \tab - \tab 9 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3544 \cr
   Total \tab 11607 \tab 15421 \tab 6113 \tab 1042 \tab 55 \tab 238 \tab 40 \tab 355 \tab 34 \tab 2612 \tab 377 \tab 151 \tab 2187 \tab 145 \tab 8 \tab 25 \tab 1642 \tab 43 \tab 1584 \tab 2 \tab 856 \tab 451 \tab 372 \tab 149 \tab 31 \tab 8 \tab 41 \tab 406 \tab 2144 \tab 3 \tab 14 \tab 17 \tab 103 \tab 123 \tab 71 \tab 49 \tab 26 \tab 10 \tab 1 \tab 1 \tab 11 \tab 14 \tab 30 \tab 19 \tab 49 \tab 17 \tab 17 \tab 1 \tab 152 \tab 10 \tab 8 \tab 3 \tab 1 \tab 3 \tab 15 \tab 7 \tab 12 \tab 9 \tab 5 \tab 1 \tab 2 \tab 4 \tab 2 \tab 12 \tab 3 \tab 3 \tab 8 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 49034 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{numwomen}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
