% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{numorg}
\alias{numorg}
\title{Number of people working in organization at all locations}
\source{
General Social Survey https://gss.norc.org
}
\description{
numorg
}
\details{
Question About how many people work in your organization at all locations? Count part-time as well as full-time employees.
}
\section{Values}{

\itemize{
\item \verb{[9981]} 1-9 in range
\item \verb{[9982]} 10-49 in range
\item \verb{[9983]} 50-99 in range
\item \verb{[9984]} 100-499 in range
\item \verb{[9985]} 500-999 in range
\item \verb{[9986]} 1,000-1,999 in range
\item \verb{[9987]} 2,000-2,999 in range
\item \verb{[9988]} 3,000-3,999 in range
\item \verb{[9989]} 4,000-4,999 in range
\item \verb{[9990]} 5,000-10,000 in range
\item \verb{[9991]} 10,000-50,000 in range
\item \verb{[9992]} 50,000-100,000 in range
\item \verb{[9993]} above 100,000 in range
\item \verb{[9997]} 10,000 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2010, 2014, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3556/vshow}{GSS Data Explorer page for \code{numorg}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 100 \tab 1000 \tab 1001 \tab 11 \tab 110 \tab 1100 \tab 113 \tab 12 \tab 120 \tab 1200 \tab 13 \tab 130 \tab 14 \tab 140 \tab 15 \tab 150 \tab 1500 \tab 154 \tab 155 \tab 16 \tab 160 \tab 1600 \tab 163 \tab 165 \tab 17 \tab 172 \tab 175 \tab 18 \tab 180 \tab 1800 \tab 19 \tab 2 \tab 20 \tab 200 \tab 2000 \tab 207 \tab 21 \tab 22 \tab 225 \tab 23 \tab 2300 \tab 234 \tab 240 \tab 25 \tab 250 \tab 2500 \tab 258 \tab 2600 \tab 27 \tab 3 \tab 30 \tab 300 \tab 3000 \tab 309 \tab 3200 \tab 33 \tab 35 \tab 350 \tab 3500 \tab 37 \tab 370 \tab 3800 \tab 4 \tab 40 \tab 400 \tab 4000 \tab 42 \tab 425 \tab 43 \tab 4300 \tab 44 \tab 45 \tab 450 \tab 467 \tab 48 \tab 49 \tab 5 \tab 50 \tab 500 \tab 5000 \tab 532 \tab 55 \tab 550 \tab 59 \tab 6 \tab 60 \tab 600 \tab 6000 \tab 61 \tab 63 \tab 630 \tab 65 \tab 6500 \tab 66 \tab 7 \tab 70 \tab 700 \tab 7000 \tab 72 \tab 745 \tab 75 \tab 750 \tab 7500 \tab 76 \tab 8 \tab 80 \tab 800 \tab 8000 \tab 83 \tab 85 \tab 8500 \tab 9 \tab 90 \tab 900 \tab 9000 \tab 92 \tab 9800 \tab 1-9 in range \tab 10-49 in range \tab 50-99 in range \tab 100-499 in range \tab 500-999 in range \tab 1,000-1,999 in range \tab 2,000-2,999 in range \tab 3,000-3,999 in range \tab 4,000-4,999 in range \tab 5,000-10,000 in range \tab 10,000-50,000 in range \tab 50,000-100,000 in range \tab above 100,000 in range \tab 10,000 or more \tab don't know \tab no answer \tab 1700 \tab 28 \tab 280 \tab 31 \tab 32 \tab 38 \tab 47 \tab 480 \tab 52 \tab 650 \tab 850 \tab 24 \tab 260 \tab 375 \tab 4500 \tab 56 \tab 73 \tab 78 \tab 99 \tab 999 \tab 105 \tab 106 \tab 116 \tab 125 \tab 1300 \tab 161 \tab 185 \tab 190 \tab 210 \tab 220 \tab 275 \tab 36 \tab 499 \tab 5100 \tab 62 \tab 670 \tab 86 \tab 95 \tab Total \cr
   2006 \tab 2778 \tab 88 \tab 22 \tab 30 \tab 34 \tab 2 \tab 9 \tab 2 \tab 1 \tab 1 \tab 15 \tab 10 \tab 5 \tab 5 \tab 3 \tab 5 \tab 1 \tab 19 \tab 20 \tab 6 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab 2 \tab 3 \tab 2 \tab 1 \tab 4 \tab 3 \tab 31 \tab 22 \tab 31 \tab 33 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 14 \tab 11 \tab 6 \tab 1 \tab 1 \tab 1 \tab 38 \tab 14 \tab 25 \tab 9 \tab 1 \tab 2 \tab 1 \tab 9 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 26 \tab 15 \tab 10 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 4 \tab 1 \tab 2 \tab 3 \tab 37 \tab 39 \tab 32 \tab 17 \tab 1 \tab 2 \tab 1 \tab 1 \tab 21 \tab 11 \tab 9 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 20 \tab 11 \tab 4 \tab 5 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 19 \tab 13 \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab 22 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 23 \tab 66 \tab 52 \tab 95 \tab 51 \tab 37 \tab 35 \tab 12 \tab 9 \tab 35 \tab 36 \tab 15 \tab 46 \tab 80 \tab 149 \tab 24 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2010 \tab 857 \tab 71 \tab 7 \tab 10 \tab 5 \tab - \tab 6 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab 8 \tab - \tab 1 \tab - \tab 5 \tab 8 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 28 \tab 10 \tab 7 \tab 8 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 22 \tab 13 \tab 4 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 13 \tab 5 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 18 \tab 11 \tab 9 \tab 7 \tab - \tab - \tab - \tab - \tab 14 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 10 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 9 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 13 \tab 4 \tab - \tab - \tab - \tab - \tab 84 \tab 110 \tab 71 \tab 114 \tab 56 \tab 28 \tab 29 \tab 15 \tab 20 \tab 38 \tab 47 \tab 21 \tab 47 \tab 26 \tab - \tab 87 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2014 \tab 1765 \tab 56 \tab 5 \tab 5 \tab 4 \tab - \tab 7 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab 5 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 29 \tab 4 \tab 6 \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 14 \tab 3 \tab 5 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 17 \tab 3 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 9 \tab 11 \tab 10 \tab 3 \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 10 \tab 1 \tab 1 \tab - \tab - \tab - \tab 30 \tab 56 \tab 28 \tab 86 \tab 29 \tab 22 \tab 15 \tab 11 \tab 11 \tab 31 \tab 29 \tab 19 \tab 46 \tab 23 \tab - \tab 42 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 62 \tab 25 \tab 36 \tab 33 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 14 \tab 7 \tab 3 \tab 8 \tab 2 \tab 1 \tab 2 \tab 17 \tab 16 \tab 7 \tab - \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 26 \tab 30 \tab 24 \tab 28 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 12 \tab 11 \tab 4 \tab - \tab - \tab 1 \tab 26 \tab 15 \tab 20 \tab 13 \tab - \tab 1 \tab - \tab 6 \tab 9 \tab 1 \tab 1 \tab 1 \tab - \tab 19 \tab 11 \tab 19 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 4 \tab - \tab - \tab - \tab 15 \tab 50 \tab 24 \tab 15 \tab - \tab - \tab 3 \tab - \tab 15 \tab 6 \tab 5 \tab 5 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 16 \tab 10 \tab 4 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab 14 \tab 4 \tab 8 \tab 3 \tab - \tab 1 \tab - \tab 13 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 9 \tab 38 \tab 31 \tab 37 \tab 30 \tab 27 \tab 16 \tab 15 \tab 9 \tab 28 \tab 29 \tab 21 \tab 67 \tab 100 \tab 124 \tab 44 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2348 \cr
   Total \tab 6300 \tab 277 \tab 59 \tab 81 \tab 76 \tab 2 \tab 26 \tab 4 \tab 3 \tab 1 \tab 37 \tab 20 \tab 10 \tab 25 \tab 7 \tab 10 \tab 3 \tab 46 \tab 49 \tab 17 \tab 1 \tab 2 \tab 14 \tab 5 \tab 1 \tab 1 \tab 3 \tab 12 \tab 2 \tab 5 \tab 9 \tab 4 \tab 6 \tab 4 \tab 114 \tab 66 \tab 68 \tab 74 \tab 1 \tab 2 \tab 5 \tab 2 \tab 4 \tab 1 \tab 1 \tab 4 \tab 29 \tab 25 \tab 11 \tab 1 \tab 1 \tab 2 \tab 100 \tab 45 \tab 54 \tab 26 \tab 1 \tab 3 \tab 1 \tab 20 \tab 12 \tab 3 \tab 2 \tab 2 \tab 2 \tab 75 \tab 34 \tab 39 \tab 10 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 10 \tab 8 \tab 1 \tab 2 \tab 6 \tab 79 \tab 111 \tab 75 \tab 42 \tab 1 \tab 2 \tab 4 \tab 1 \tab 56 \tab 22 \tab 16 \tab 12 \tab 1 \tab 3 \tab 1 \tab 10 \tab 3 \tab 1 \tab 51 \tab 26 \tab 9 \tab 9 \tab 2 \tab 1 \tab 15 \tab 2 \tab 2 \tab 1 \tab 43 \tab 18 \tab 18 \tab 6 \tab 2 \tab 2 \tab 1 \tab 58 \tab 8 \tab 4 \tab 4 \tab 1 \tab 1 \tab 146 \tab 270 \tab 182 \tab 332 \tab 166 \tab 114 \tab 95 \tab 53 \tab 49 \tab 132 \tab 141 \tab 76 \tab 206 \tab 229 \tab 273 \tab 197 \tab 3 \tab 2 \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 11440 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
