% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{numorg1}
\alias{numorg1}
\title{Estimate of total people in r's business or organization}
\source{
General Social Survey https://gss.norc.org
}
\description{
numorg1
}
\details{
Question About how many people work in your organization at all locations? Count part-time as well as full-time employees.
}
\section{Values}{

\itemize{
\item \code{1} 1-9 in range
\item \code{2} 10-49 in range
\item \code{3} 50-99 in range
\item \code{4} 100-499 in range
\item \code{5} 500-999 in range
\item \code{6} 1,000-1,999 in range
\item \code{7} 2,000-2,999 in range
\item \code{8} 3,000-3,999 in range
\item \code{9} 4,000-4,999 in range
\item \code{10} 5,000-10,000 in range
\item \code{11} 10,000-50,000 in range
\item \code{12} 50,000-100,000 in range
\item \code{13} above 100,000 in range
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7852/vshow}{GSS Data Explorer page for \code{numorg1}}.
}

\keyword{variable}
