% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_317.R
\name{numpets}
\alias{numpets}
\title{How many pets r's family has}
\source{
General Social Survey https://gss.norc.org
}
\description{
numpets
}
\details{
Question How many pets does your family have? By family, we mean members of your family with whom you currently live.
}
\section{Values}{

\itemize{
\item \code{0} no pets
\item \code{20} 20 or more pets
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7194/vshow}{GSS Data Explorer page for \code{numpets}}.
}

\keyword{variable}
