% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{numprobs}
\alias{numprobs}
\title{How many friends close to discuss problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
numprobs
}
\details{
Question Of these (NUMCNTCT) friends and relatives, about how many would you say you feel really close to, that is close enough to discuss personal or important problems with?
}
\section{Values}{

\itemize{
\item \verb{[96]} 96 or higher
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2606/vshow}{GSS Data Explorer page for \code{numprobs}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 24 \tab 25 \tab 26 \tab 3 \tab 30 \tab 34 \tab 35 \tab 38 \tab 4 \tab 40 \tab 5 \tab 50 \tab 55 \tab 6 \tab 60 \tab 65 \tab 7 \tab 70 \tab 75 \tab 8 \tab 9 \tab 96 or higher \tab don't know \tab no answer \tab Total \cr
   2002 \tab 74 \tab 101 \tab 200 \tab 304 \tab 17 \tab 41 \tab 7 \tab 2 \tab 129 \tab 7 \tab 5 \tab 375 \tab 100 \tab 2 \tab 68 \tab 2 \tab 366 \tab 31 \tab 1 \tab 6 \tab 1 \tab 199 \tab 8 \tab 352 \tab 35 \tab 1 \tab 163 \tab 4 \tab 1 \tab 48 \tab 1 \tab 7 \tab 70 \tab 9 \tab 10 \tab 4 \tab 14 \tab 2765 \cr
   Total \tab 74 \tab 101 \tab 200 \tab 304 \tab 17 \tab 41 \tab 7 \tab 2 \tab 129 \tab 7 \tab 5 \tab 375 \tab 100 \tab 2 \tab 68 \tab 2 \tab 366 \tab 31 \tab 1 \tab 6 \tab 1 \tab 199 \tab 8 \tab 352 \tab 35 \tab 1 \tab 163 \tab 4 \tab 1 \tab 48 \tab 1 \tab 7 \tab 70 \tab 9 \tab 10 \tab 4 \tab 14 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{byemprob}},
\code{\link{intaccss}},
\code{\link{jobinfo}},
\code{\link{jobpaper}}
}
\concept{Information Society}
\keyword{variable}
