% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{numpromo}
\alias{numpromo}
\title{#Promotions have you received}
\source{
General Social Survey https://gss.norc.org
}
\description{
numpromo
}
\details{
Question Have you received any promotions while working for your present employer (IF SELF-EMPLOYED:  current organization)?
A. How many promotions have you received?
}
\section{Values}{

\itemize{
\item \verb{[7]} 7 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1301/vshow}{GSS Data Explorer page for \code{numpromo}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 or more \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab 1148 \tab 132 \tab 61 \tab 41 \tab 26 \tab 5 \tab 3 \tab 10 \tab 1 \tab 90 \tab - \tab 1517 \cr
   Total \tab 1148 \tab 132 \tab 61 \tab 41 \tab 26 \tab 5 \tab 3 \tab 10 \tab 1 \tab 90 \tab 0 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{afailure}},
\code{\link{automatn}},
\code{\link{baseofop}},
\code{\link{chngeorg}},
\code{\link{cowrkrel}},
\code{\link{drugtest}},
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{futpromo}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{helporg}},
\code{\link{idecide}},
\code{\link{imatter}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobcntrl}},
\code{\link{jobjudge}},
\code{\link{jobvshme}},
\code{\link{jobyears}},
\code{\link{learning}},
\code{\link{levels}()},
\code{\link{lottosay}},
\code{\link{ltrsref}},
\code{\link{merged}},
\code{\link{monitred}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{numunemp}},
\code{\link{offsup}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{orgfin}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{othpay}},
\code{\link{outofbiz}},
\code{\link{pessimst}},
\code{\link{physical}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{proudorg}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{racpromo}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{reorg}},
\code{\link{samejob}},
\code{\link{samevals}},
\code{\link{satself}},
\code{\link{schoolng}},
\code{\link{sexpromo}},
\code{\link{skiltest}},
\code{\link{slfrspct}},
\code{\link{splocnum}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{totunemp}},
\code{\link{training}},
\code{\link{unmanrel}},
\code{\link{unpower}},
\code{\link{unprog}},
\code{\link{unvote}},
\code{\link{wherewrk}},
\code{\link{wrkindep}},
\code{\link{wrkmuch}},
\code{\link{wrkwell}},
\code{\link{wrkyears}},
\code{\link{youmoney}},
\code{\link{yousup}}

Other Employment: 
\code{\link{accntsci}},
\code{\link{advances}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compwage}},
\code{\link{curempyr}},
\code{\link{denyrais}},
\code{\link{dowell}},
\code{\link{drugtest}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empyears}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{evunemp}},
\code{\link{expernc}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findjob}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{futpromo}},
\code{\link{harjob5}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{ignorwk}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobyears}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{knowschd}},
\code{\link{laidoff}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leavejb}},
\code{\link{liedcwkr}},
\code{\link{losejb12}},
\code{\link{losejob5}},
\code{\link{lostjobs}},
\code{\link{ltrsref}},
\code{\link{marrcoun}},
\code{\link{medtreat}},
\code{\link{moredays}},
\code{\link{moretrde}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{neverwk}},
\code{\link{newjobs}},
\code{\link{numemp}},
\code{\link{numemply}},
\code{\link{numemps}},
\code{\link{numemps5}},
\code{\link{numemps_0422}},
\code{\link{numown}},
\code{\link{numunemp}},
\code{\link{ownbiz}},
\code{\link{paidemps}},
\code{\link{partfull}},
\code{\link{paysame}},
\code{\link{physical}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{ryrsfirm}},
\code{\link{samejob}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{skiltest}},
\code{\link{slsmnshp}},
\code{\link{spwrkgvt}},
\code{\link{startbiz}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totunemp}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{whywkhme}},
\code{\link{wkbhvrs}},
\code{\link{work3yrs}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrkstat}},
\code{\link{wrktype}},
\code{\link{wrkyears}},
\code{\link{yearbiz}},
\code{\link{yearsjob}},
\code{\link{yrsfirm}}
}
\concept{Employment}
\concept{Work Organizations}
\keyword{variable}
