% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{numrelex}
\alias{numrelex}
\title{Number of real experience that changed r's life}
\source{
General Social Survey https://gss.norc.org
}
\description{
numrelex
}
\details{
Question IF RELEXPER = yes and RELEXP = no, READ: Thinking of the "turning point in your life when you made a new and personal commitment to religion," how many such experiences have you had?
READ IF NEEDED: THAT IS ON HOW MANY DIFFERENT OCCASIONS DID THIS OCCUR?
IF RELEXPER = no and RELEXP = yes, READ: Thinking of the "religious or spiritual experience that changed your life," how many such experiences have you had?
READ IF NEEDED: THAT IS ON HOW MANY DIFFERENT OCCASIONS DID THIS OCCUR?
IF RELEXPER = yes and RELEXP = yes, READ: Thinking of the "turning point in your life when you made a new and personal commitment to religion," and the "religious or spiritual experience that changed your life," how many such experiences have you had?
READ IF NEEDED: THAT IS ON HOW MANY DIFFERENT OCCASIONS DID THIS OCCUR?
}
\section{Values}{

\itemize{
\item \verb{[7]} seven or more
\item \verb{[8]} 1 or more experience, dk
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3061/vshow}{GSS Data Explorer page for \code{numrelex}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab seven or more \tab 1 or more experience, dk \tab no answer \tab Total \cr
   2004 \tab 2178 \tab 318 \tab 142 \tab 61 \tab 20 \tab 14 \tab 6 \tab 30 \tab 33 \tab 10 \tab 2812 \cr
   Total \tab 2178 \tab 318 \tab 142 \tab 61 \tab 20 \tab 14 \tab 6 \tab 30 \tab 33 \tab 10 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{agerelex}},
\code{\link{chngrel}}
}
\concept{Spirituality and Religion}
\keyword{variable}
