% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{numrooms}
\alias{numrooms}
\title{Number of rooms in home}
\source{
General Social Survey https://gss.norc.org
}
\description{
numrooms
}
\details{
Question How many rooms are there in your home (apartment or house)? Do not count any separate kitchens, bathrooms, garages, balconies, hallways or cupboards.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7500/vshow}{GSS Data Explorer page for \code{numrooms}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 3 \tab 83 \tab 63 \tab 24 \tab 28 \tab 10 \tab 9 \tab 3 \tab 3 \tab 1 \tab 189 \tab 1 \tab 270 \tab 221 \tab 259 \tab 235 \tab 165 \tab 156 \tab 89 \tab 192 \tab 25 \tab 4032 \cr
   Total \tab 2003 \tab 3 \tab 83 \tab 63 \tab 24 \tab 28 \tab 10 \tab 9 \tab 3 \tab 3 \tab 1 \tab 189 \tab 1 \tab 270 \tab 221 \tab 259 \tab 235 \tab 165 \tab 156 \tab 89 \tab 192 \tab 25 \tab 4032 \cr
}
}

\keyword{variable}
