% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{numsites}
\alias{numsites}
\title{Does r's firm have more than one location}
\source{
General Social Survey https://gss.norc.org
}
\description{
numsites
}
\details{
Question Does your company or the institution for which you work have more than one physical location (that is, other offices, branches, plants, stores, or subsidiaries)? (If you work for the government, does your agency or division have more than one physical location?)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4566/vshow}{GSS Data Explorer page for \code{numsites}}.
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numemply}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{totalnum}}

Other Job: 
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{totalnum}}
}
\concept{Job}
\concept{Work Orientation}
\keyword{variable}
