% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{numunemp}
\alias{numunemp}
\title{How many times unemployed more than two weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
numunemp
}
\details{
Question About how many different times have you been unemployed and looking for work for as long as two weeks?
}
\section{Values}{

\itemize{
\item \verb{[96]} 96 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1273/vshow}{GSS Data Explorer page for \code{numunemp}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 12 \tab 15 \tab 2 \tab 20 \tab 21 \tab 25 \tab 3 \tab 30 \tab 32 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 70 \tab 8 \tab 96 or more \tab don't know \tab no answer \tab Total \cr
   1991 \tab 824 \tab 211 \tab 27 \tab 6 \tab 4 \tab 156 \tab 6 \tab 1 \tab 2 \tab 114 \tab 3 \tab 1 \tab 66 \tab 35 \tab 1 \tab 17 \tab 7 \tab 1 \tab 5 \tab 2 \tab 12 \tab 16 \tab 1517 \cr
   Total \tab 824 \tab 211 \tab 27 \tab 6 \tab 4 \tab 156 \tab 6 \tab 1 \tab 2 \tab 114 \tab 3 \tab 1 \tab 66 \tab 35 \tab 1 \tab 17 \tab 7 \tab 1 \tab 5 \tab 2 \tab 12 \tab 16 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{orgfin}},
\code{\link{totunemp}},
\code{\link{wrkyears}}
}
\concept{Work Organizations}
\keyword{variable}
