% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{numwomen}
\alias{numwomen}
\title{Number of female sex partner's since 18}
\source{
General Social Survey https://gss.norc.org
}
\description{
numwomen
}
\details{
Question Now thinking about the time since your 18th birthday (including the past 12 months) how many female partners have you had sex with?
}
\section{Values}{

\itemize{
\item \verb{[989]} 989 or more
\item \verb{[990]} dash or slash
\item \verb{[991]} 1 or more, number unknown
\item \verb{[992]} 'x'
\item \verb{[993]} garbled text
\item \verb{[994]} 'several'
\item \verb{[995]} 'many,' 'lots'
\item \verb{[996]} 'n.a.'
\item \verb{[997]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5060/vshow}{GSS Data Explorer page for \code{numwomen}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 125 \tab 13 \tab 14 \tab 15 \tab 16 \tab 165 \tab 17 \tab 175 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 253 \tab 27 \tab 28 \tab 3 \tab 30 \tab 300 \tab 35 \tab 36 \tab 4 \tab 40 \tab 45 \tab 47 \tab 5 \tab 50 \tab 56 \tab 6 \tab 7 \tab 70 \tab 75 \tab 8 \tab 80 \tab 9 \tab dash or slash \tab garbled text \tab 'several' \tab 'many,' 'lots' \tab don't know \tab no answer \tab 130 \tab 150 \tab 19 \tab 200 \tab 33 \tab refused \tab 101 \tab 11 \tab 145 \tab 34 \tab 68 \tab 90 \tab 1 or more, number unknown \tab 'x' \tab 'n.a.' \tab 120 \tab 250 \tab 26 \tab 32 \tab 403 \tab 60 \tab 167 \tab 42 \tab 450 \tab 500 \tab 52 \tab 53 \tab 61 \tab 63 \tab 65 \tab 89 \tab 989 or more \tab 137 \tab 138 \tab 147 \tab 210 \tab 270 \tab 280 \tab 29 \tab 66 \tab 85 \tab 37 \tab 41 \tab 59 \tab 350 \tab 86 \tab 140 \tab 325 \tab 38 \tab 39 \tab 55 \tab 57 \tab 77 \tab 82 \tab 103 \tab 123 \tab 48 \tab 700 \tab 73 \tab 88 \tab 105 \tab 121 \tab 122 \tab 201 \tab 301 \tab 31 \tab 43 \tab 51 \tab 54 \tab 69 \tab 170 \tab 240 \tab 46 \tab 62 \tab 127 \tab 400 \tab 49 \tab 58 \tab 750 \tab 222 \tab 74 \tab 87 \tab 84 \tab 115 \tab 255 \tab 380 \tab 96 \tab 110 \tab 132 \tab 205 \tab 311 \tab 76 \tab 420 \tab 95 \tab Total \cr
   1989 \tab 136 \tab 678 \tab 133 \tab 31 \tab 12 \tab 18 \tab 1 \tab 1 \tab 2 \tab 20 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 39 \tab 26 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 1 \tab 1 \tab 1 \tab 52 \tab 12 \tab 1 \tab 1 \tab 1 \tab 37 \tab 8 \tab 1 \tab 1 \tab 48 \tab 8 \tab 1 \tab 35 \tab 7 \tab 1 \tab 2 \tab 11 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab 16 \tab 155 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 199 \tab 562 \tab 98 \tab 30 \tab 6 \tab 11 \tab - \tab 1 \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab 2 \tab 52 \tab 33 \tab - \tab 1 \tab - \tab 2 \tab 10 \tab - \tab 1 \tab - \tab 45 \tab 14 \tab - \tab 1 \tab - \tab 44 \tab 7 \tab 1 \tab - \tab 22 \tab 16 \tab - \tab 17 \tab 9 \tab - \tab - \tab 15 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 8 \tab 145 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 221 \tab 688 \tab 94 \tab 31 \tab 8 \tab 12 \tab - \tab 2 \tab 2 \tab 8 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 53 \tab 31 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab 32 \tab 13 \tab 2 \tab 2 \tab 1 \tab 43 \tab 4 \tab - \tab - \tab 38 \tab 10 \tab - \tab 29 \tab 11 \tab - \tab 4 \tab 12 \tab - \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab 104 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 114 \tab 798 \tab 128 \tab 45 \tab 11 \tab 19 \tab - \tab 2 \tab - \tab 21 \tab 2 \tab - \tab - \tab - \tab 5 \tab 56 \tab 41 \tab 1 \tab 2 \tab 1 \tab 2 \tab 15 \tab - \tab - \tab 1 \tab 51 \tab 10 \tab 2 \tab 5 \tab - \tab 26 \tab 5 \tab 1 \tab - \tab 39 \tab 10 \tab - \tab 29 \tab 16 \tab - \tab 2 \tab 14 \tab - \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 11 \tab 92 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 201 \tab 1373 \tab 208 \tab 85 \tab 17 \tab 27 \tab 1 \tab 1 \tab 3 \tab 49 \tab 3 \tab - \tab 5 \tab - \tab 3 \tab 94 \tab 57 \tab - \tab - \tab - \tab 1 \tab 25 \tab - \tab - \tab 4 \tab 90 \tab 28 \tab 1 \tab 9 \tab - \tab 73 \tab 8 \tab - \tab - \tab 70 \tab 33 \tab - \tab 55 \tab 24 \tab 1 \tab 2 \tab 32 \tab - \tab 8 \tab 5 \tab 4 \tab 3 \tab 6 \tab 34 \tab 270 \tab - \tab 3 \tab - \tab 4 \tab 4 \tab 23 \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 11 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 258 \tab 1278 \tab 180 \tab 97 \tab 20 \tab 35 \tab - \tab 8 \tab 2 \tab 44 \tab 3 \tab - \tab 1 \tab 1 \tab 6 \tab 102 \tab 50 \tab - \tab - \tab 1 \tab 6 \tab 23 \tab - \tab - \tab 1 \tab 93 \tab 22 \tab 3 \tab 4 \tab - \tab 69 \tab 6 \tab - \tab - \tab 69 \tab 18 \tab - \tab 53 \tab 32 \tab 6 \tab 3 \tab 39 \tab 2 \tab 7 \tab 5 \tab 1 \tab 2 \tab 19 \tab 17 \tab 222 \tab 2 \tab 4 \tab 1 \tab 3 \tab - \tab 24 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 35 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 383 \tab 1254 \tab 186 \tab 85 \tab 12 \tab 27 \tab - \tab 2 \tab 1 \tab 41 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 81 \tab 66 \tab - \tab - \tab 1 \tab 1 \tab 17 \tab - \tab - \tab - \tab 81 \tab 23 \tab 2 \tab 8 \tab 1 \tab 69 \tab 10 \tab 1 \tab - \tab 80 \tab 25 \tab - \tab 52 \tab 28 \tab 2 \tab 1 \tab 30 \tab - \tab 5 \tab 2 \tab - \tab 2 \tab 6 \tab 24 \tab 166 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 9 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 13 \tab - \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 417 \tab 1118 \tab 188 \tab 73 \tab 17 \tab 24 \tab - \tab 1 \tab 4 \tab 28 \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 81 \tab 44 \tab 2 \tab 2 \tab 4 \tab - \tab 19 \tab - \tab 2 \tab 2 \tab 89 \tab 24 \tab 2 \tab 2 \tab - \tab 80 \tab 10 \tab 3 \tab 1 \tab 69 \tab 30 \tab - \tab 56 \tab 25 \tab 2 \tab 2 \tab 32 \tab - \tab 9 \tab 1 \tab - \tab - \tab 2 \tab 25 \tab 255 \tab - \tab 2 \tab 1 \tab 7 \tab - \tab 23 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 490 \tab 1198 \tab 145 \tab 78 \tab 19 \tab 24 \tab 1 \tab 5 \tab 5 \tab 28 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 79 \tab 55 \tab 3 \tab - \tab 1 \tab 2 \tab 19 \tab - \tab 2 \tab 3 \tab 67 \tab 24 \tab 2 \tab 10 \tab 1 \tab 46 \tab 10 \tab - \tab 1 \tab 84 \tab 12 \tab - \tab 60 \tab 20 \tab 1 \tab 3 \tab 24 \tab 2 \tab 17 \tab - \tab - \tab - \tab 1 \tab 49 \tab 130 \tab - \tab 4 \tab - \tab 4 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 584 \tab 1177 \tab 191 \tab 63 \tab 20 \tab 27 \tab - \tab 3 \tab 5 \tab 29 \tab 3 \tab - \tab 3 \tab - \tab 5 \tab 75 \tab 49 \tab 4 \tab 1 \tab 3 \tab 3 \tab 25 \tab - \tab - \tab - \tab 83 \tab 20 \tab 1 \tab 9 \tab 1 \tab 50 \tab 8 \tab 2 \tab - \tab 72 \tab 21 \tab 1 \tab 34 \tab 25 \tab 2 \tab 2 \tab 22 \tab - \tab 8 \tab - \tab - \tab 1 \tab 6 \tab 41 \tab 86 \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 11 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 6 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2096 \tab 1293 \tab 207 \tab 91 \tab 16 \tab 23 \tab - \tab 6 \tab 8 \tab 39 \tab 3 \tab - \tab 2 \tab - \tab 8 \tab 87 \tab 41 \tab 3 \tab 2 \tab 1 \tab 4 \tab 18 \tab - \tab - \tab 1 \tab 86 \tab 19 \tab 1 \tab 4 \tab 3 \tab 64 \tab 13 \tab 2 \tab - \tab 83 \tab 15 \tab 1 \tab 51 \tab 22 \tab 1 \tab 1 \tab 29 \tab 3 \tab 10 \tab - \tab - \tab - \tab 6 \tab 33 \tab 75 \tab - \tab - \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 240 \tab 912 \tab 140 \tab 44 \tab 15 \tab 26 \tab - \tab 5 \tab 6 \tab 20 \tab 1 \tab - \tab 4 \tab - \tab 4 \tab 74 \tab 33 \tab - \tab 4 \tab 2 \tab 3 \tab 22 \tab - \tab 1 \tab - \tab 67 \tab 16 \tab 2 \tab - \tab 1 \tab 54 \tab 11 \tab 2 \tab - \tab 61 \tab 16 \tab - \tab 30 \tab 18 \tab - \tab 3 \tab 32 \tab 1 \tab 6 \tab - \tab 1 \tab - \tab 2 \tab 52 \tab 61 \tab - \tab 5 \tab 3 \tab 4 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 202 \tab 956 \tab 142 \tab 64 \tab 17 \tab 22 \tab - \tab 5 \tab 6 \tab 23 \tab 7 \tab - \tab 3 \tab - \tab - \tab 66 \tab 43 \tab 3 \tab 5 \tab - \tab 1 \tab 16 \tab - \tab 1 \tab 1 \tab 66 \tab 19 \tab - \tab 9 \tab 2 \tab 40 \tab 11 \tab 3 \tab - \tab 64 \tab 18 \tab - \tab 42 \tab 9 \tab 2 \tab - \tab 15 \tab 2 \tab 12 \tab - \tab - \tab 2 \tab 2 \tab 41 \tab 67 \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 6 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 238 \tab 888 \tab 150 \tab 48 \tab 8 \tab 17 \tab - \tab 3 \tab 2 \tab 26 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 60 \tab 32 \tab 2 \tab 1 \tab 1 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab 61 \tab 14 \tab - \tab 5 \tab - \tab 52 \tab 12 \tab 3 \tab - \tab 51 \tab 11 \tab - \tab 37 \tab 26 \tab 1 \tab 3 \tab 23 \tab - \tab 14 \tab - \tab 1 \tab - \tab 2 \tab 50 \tab 87 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 188 \tab 1183 \tab 204 \tab 80 \tab 15 \tab 32 \tab 1 \tab 6 \tab 1 \tab 35 \tab 3 \tab - \tab 3 \tab - \tab 4 \tab 96 \tab 53 \tab 2 \tab 3 \tab 1 \tab 2 \tab 26 \tab - \tab 1 \tab - \tab 92 \tab 26 \tab 2 \tab 4 \tab 2 \tab 59 \tab 10 \tab 2 \tab - \tab 72 \tab 27 \tab - \tab 41 \tab 23 \tab 1 \tab - \tab 31 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab - \tab 115 \tab 47 \tab - \tab 4 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 11 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1096 \tab 880 \tab 156 \tab 51 \tab 9 \tab 24 \tab - \tab 6 \tab 6 \tab 23 \tab 5 \tab - \tab 2 \tab - \tab 6 \tab 85 \tab 37 \tab 2 \tab - \tab 2 \tab 2 \tab 18 \tab - \tab 2 \tab 1 \tab 68 \tab 14 \tab 1 \tab 6 \tab - \tab 54 \tab 7 \tab - \tab - \tab 60 \tab 12 \tab - \tab 46 \tab 18 \tab - \tab 3 \tab 26 \tab - \tab 13 \tab - \tab - \tab - \tab 2 \tab 21 \tab 68 \tab - \tab 3 \tab 2 \tab 6 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 942 \tab 670 \tab 133 \tab 35 \tab 3 \tab 20 \tab - \tab 4 \tab 3 \tab 28 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 68 \tab 33 \tab 1 \tab - \tab 1 \tab - \tab 9 \tab - \tab - \tab - \tab 55 \tab 20 \tab - \tab 3 \tab - \tab 43 \tab 8 \tab - \tab - \tab 62 \tab 16 \tab - \tab 30 \tab 11 \tab - \tab 2 \tab 24 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 30 \tab 50 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 9 \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1703 \tab 1160 \tab 237 \tab 74 \tab 11 \tab 34 \tab - \tab 1 \tab 4 \tab 28 \tab 5 \tab - \tab 1 \tab - \tab 4 \tab 108 \tab 44 \tab 1 \tab 1 \tab 4 \tab 3 \tab 14 \tab - \tab - \tab 1 \tab 79 \tab 17 \tab 1 \tab 5 \tab 2 \tab 69 \tab 8 \tab - \tab - \tab 90 \tab 26 \tab - \tab 51 \tab 28 \tab 2 \tab 5 \tab 24 \tab 1 \tab 9 \tab - \tab 1 \tab - \tab 6 \tab 19 \tab 103 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab 7 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4032 \cr
   2022 \tab 1898 \tab 741 \tab 143 \tab 37 \tab 9 \tab 9 \tab - \tab 2 \tab 2 \tab 18 \tab 3 \tab - \tab 2 \tab - \tab - \tab 53 \tab 19 \tab - \tab 3 \tab 2 \tab 1 \tab 11 \tab - \tab - \tab 2 \tab 57 \tab 9 \tab - \tab 2 \tab - \tab 265 \tab 5 \tab 3 \tab 1 \tab 52 \tab 4 \tab - \tab 34 \tab 15 \tab - \tab 2 \tab 25 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 89 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 11606 \tab 18807 \tab 3063 \tab 1142 \tab 245 \tab 431 \tab 4 \tab 64 \tab 64 \tab 515 \tab 51 \tab 1 \tab 36 \tab 3 \tab 70 \tab 1409 \tab 787 \tab 26 \tab 27 \tab 27 \tab 35 \tab 312 \tab 1 \tab 11 \tab 21 \tab 1314 \tab 344 \tab 23 \tab 89 \tab 15 \tab 1237 \tab 161 \tab 24 \tab 4 \tab 1186 \tab 328 \tab 3 \tab 782 \tab 367 \tab 22 \tab 40 \tab 460 \tab 14 \tab 148 \tab 18 \tab 12 \tab 14 \tab 68 \tab 594 \tab 2272 \tab 3 \tab 39 \tab 20 \tab 59 \tab 11 \tab 99 \tab 5 \tab 85 \tab 1 \tab 6 \tab 2 \tab 11 \tab 99 \tab 9 \tab 41 \tab 14 \tab 7 \tab 10 \tab 13 \tab 1 \tab 49 \tab 1 \tab 5 \tab 1 \tab 9 \tab 5 \tab 1 \tab 1 \tab 3 \tab 8 \tab 1 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 5 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 49034 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
