% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_88.R
\name{nurse}
\alias{nurse}
\title{Women as nurses in combat zone}
\source{
General Social Survey https://gss.norc.org
}
\description{
nurse
}
\details{
Question I'm going to read you a list of jobs that people might have in the armed forces. Please tell me whether you think a woman should or should not be assigned to each job, assuming she is trained to do it.
C. A nurse in a combat zone ...
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/791/vshow}{GSS Data Explorer page for \code{nurse}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 1676 \tab 145 \tab 28 \tab 11 \tab 1860 \cr
   Total \tab 0 \tab 1676 \tab 145 \tab 28 \tab 11 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{brass}},
\code{\link{fedirty}},
\code{\link{fefight}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{gunner}},
\code{\link{mechanic}},
\code{\link{transair}},
\code{\link{typist}}
}
\concept{Core}
\keyword{variable}
