% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_92.R
\name{ob911}
\alias{ob911}
\title{Obligation--reporting a crime}
\source{
General Social Survey https://gss.norc.org
}
\description{
ob911
}
\details{
Question And one last question. We all know that American citizens have certain rights. For example, they have the right to free public education and to police protection, the right to attend religious services of their choice, and the right to elec public officials. I'd like to ask now about certain obligations that some people feel American citizens owe their country. I just want your own opinion on these--whether you feel it is a very important obligation, a somewhat important obligation, or not an obligation that a citizen owes to the country. (READ EACH STATEMENT AND CIRCLE ONE CODE for each).
(REPEAT ANSWER CATEGORIES AS NECESSARY)
D. Reporting a crime that he or she may have witnessed?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} somewhat important
\item \verb{[3]} not obligation
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/837/vshow}{GSS Data Explorer page for \code{ob911}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very important \tab somewhat important \tab not obligation \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 1312 \tab 114 \tab 16 \tab 8 \tab 23 \tab 1473 \cr
   Total \tab 0 \tab 1312 \tab 114 \tab 16 \tab 8 \tab 23 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{defwrknw}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{obeng}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{resdefwk}}
}
\concept{Core}
\keyword{variable}
