% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{obeylaws}
\alias{obeylaws}
\title{How important always to obey laws}
\source{
General Social Survey https://gss.norc.org
}
\description{
obeylaws
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
C. Always to obey laws and regulations
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4909/vshow}{GSS Data Explorer page for \code{obeylaws}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 5 \tab 2 \tab 18 \tab 52 \tab 141 \tab 263 \tab 986 \tab 3 \tab 2 \tab 2812 \cr
   2014 \tab 1274 \tab 7 \tab 10 \tab 14 \tab 50 \tab 94 \tab 222 \tab 853 \tab 12 \tab 2 \tab 2538 \cr
   Total \tab 2614 \tab 12 \tab 12 \tab 32 \tab 102 \tab 235 \tab 485 \tab 1839 \tab 15 \tab 4 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{relmeet}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
