% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_56.R
\name{obeysy}
\alias{obeysy}
\title{Obeys parents well-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
obeysy
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
9. That a child obeys his or her parents well
}
\section{Values}{

\itemize{
\item \code{1} 1 most desirable
\item \code{2} 3 most desirable
\item \code{3} not mentioned
\item \code{4} 3 least desirable
\item \code{5} 1 least desirable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1980 \tab - \tab full \cr
   1984 \tab - \tab full \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/495/vshow}{GSS Data Explorer page for \code{obeysy}}.
}

\seealso{
Other Core: 
\code{\link{cleanz}},
\code{\link{considey}},
\code{\link{honestz}},
\code{\link{interesy}},
\code{\link{judgmenz}},
\code{\link{mannersz}},
\code{\link{responsy}},
\code{\link{studiouy}},
\code{\link{successz}}
}
\concept{Core}
\keyword{variable}
