% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{obmewar}
\alias{obmewar}
\title{Obligation--wartime mil service-men}
\source{
General Social Survey https://gss.norc.org
}
\description{
obmewar
}
\details{
Question And one last question. We all know that American citizens have certain rights. For example, they have the right to free public education and to police protection, the right to attend religious services of their choice, and the right to elec public officials. I'd like to ask now about certain obligations that some people feel American citizens owe their country. I just want your own opinion on these--whether you feel it is a very important obligation, a somewhat important obligation, or not an obligation that a citizen owes to the country. (READ EACH STATEMENT AND CIRCLE ONE CODE for each).
(REPEAT ANSWER CATEGORIES AS NECESSARY)
H. For young men, serving in the military when the country is at war?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} somewhat important
\item \code{3} not obligation
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/841/vshow}{GSS Data Explorer page for \code{obmewar}}.
}

\seealso{
Other Core: 
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{obmepax}}

Other Citizenship: 
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{obmepax}}
}
\concept{Citizenship}
\concept{Core}
\keyword{variable}
