% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{occfirst}
\alias{occfirst}
\title{R first occupation}
\source{
General Social Survey https://gss.norc.org
}
\description{
occfirst
}
\details{
Question Respondent's First Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1462/vshow}{GSS Data Explorer page for \code{occfirst}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 104 \tab 118 \tab 127 \tab 128 \tab 13 \tab 136 \tab 14 \tab 143 \tab 15 \tab 154 \tab 156 \tab 157 \tab 158 \tab 159 \tab 16 \tab 163 \tab 164 \tab 165 \tab 167 \tab 174 \tab 176 \tab 177 \tab 178 \tab 18 \tab 185 \tab 186 \tab 187 \tab 188 \tab 19 \tab 193 \tab 194 \tab 197 \tab 198 \tab 199 \tab 203 \tab 206 \tab 207 \tab 208 \tab 213 \tab 216 \tab 217 \tab 218 \tab 224 \tab 225 \tab 226 \tab 228 \tab 229 \tab 23 \tab 234 \tab 235 \tab 243 \tab 25 \tab 253 \tab 254 \tab 256 \tab 259 \tab 263 \tab 264 \tab 265 \tab 266 \tab 267 \tab 268 \tab 269 \tab 27 \tab 274 \tab 275 \tab 276 \tab 277 \tab 278 \tab 303 \tab 307 \tab 308 \tab 313 \tab 315 \tab 316 \tab 317 \tab 318 \tab 319 \tab 327 \tab 328 \tab 329 \tab 33 \tab 335 \tab 336 \tab 337 \tab 338 \tab 339 \tab 343 \tab 344 \tab 347 \tab 348 \tab 354 \tab 355 \tab 356 \tab 357 \tab 359 \tab 363 \tab 364 \tab 365 \tab 368 \tab 373 \tab 375 \tab 376 \tab 378 \tab 379 \tab 383 \tab 384 \tab 385 \tab 387 \tab 389 \tab 403 \tab 404 \tab 405 \tab 406 \tab 407 \tab 415 \tab 417 \tab 418 \tab 423 \tab 426 \tab 427 \tab 43 \tab 430 \tab 433 \tab 434 \tab 435 \tab 436 \tab 438 \tab 439 \tab 44 \tab 443 \tab 444 \tab 445 \tab 446 \tab 447 \tab 449 \tab 453 \tab 454 \tab 456 \tab 458 \tab 459 \tab 464 \tab 465 \tab 466 \tab 468 \tab 473 \tab 475 \tab 479 \tab 48 \tab 486 \tab 487 \tab 488 \tab 5 \tab 503 \tab 505 \tab 507 \tab 508 \tab 514 \tab 518 \tab 523 \tab 525 \tab 527 \tab 53 \tab 533 \tab 534 \tab 538 \tab 55 \tab 554 \tab 56 \tab 565 \tab 566 \tab 567 \tab 569 \tab 57 \tab 573 \tab 575 \tab 576 \tab 585 \tab 587 \tab 588 \tab 59 \tab 595 \tab 599 \tab 614 \tab 63 \tab 633 \tab 634 \tab 635 \tab 636 \tab 637 \tab 639 \tab 64 \tab 643 \tab 645 \tab 647 \tab 653 \tab 666 \tab 67 \tab 678 \tab 683 \tab 686 \tab 687 \tab 688 \tab 689 \tab 7 \tab 704 \tab 706 \tab 709 \tab 715 \tab 723 \tab 725 \tab 727 \tab 734 \tab 737 \tab 738 \tab 739 \tab 744 \tab 745 \tab 747 \tab 748 \tab 749 \tab 754 \tab 755 \tab 756 \tab 759 \tab 76 \tab 766 \tab 769 \tab 773 \tab 774 \tab 777 \tab 779 \tab 78 \tab 783 \tab 785 \tab 79 \tab 796 \tab 797 \tab 799 \tab 8 \tab 804 \tab 806 \tab 829 \tab 84 \tab 85 \tab 856 \tab 859 \tab 86 \tab 863 \tab 865 \tab 869 \tab 877 \tab 878 \tab 883 \tab 885 \tab 887 \tab 888 \tab 889 \tab 9 \tab 95 \tab 96 \tab 98 \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1587 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 6 \tab 3 \tab 39 \tab 13 \tab 4 \tab 1 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 10 \tab 2 \tab 1 \tab 7 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 22 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 7 \tab 2 \tab 2 \tab 2 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 18 \tab 6 \tab 2 \tab 14 \tab 1 \tab 7 \tab 2 \tab 2 \tab 12 \tab 3 \tab 18 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab 27 \tab 3 \tab 61 \tab 3 \tab 2 \tab 2 \tab 1 \tab 5 \tab 65 \tab 10 \tab 2 \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab 3 \tab 23 \tab 4 \tab 6 \tab 1 \tab 2 \tab 1 \tab 13 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 6 \tab 4 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 29 \tab 8 \tab 1 \tab 7 \tab 3 \tab 12 \tab 2 \tab 1 \tab 1 \tab 6 \tab 13 \tab 1 \tab 1 \tab 10 \tab 2 \tab 6 \tab 3 \tab 2 \tab 7 \tab 3 \tab 6 \tab 48 \tab 27 \tab 5 \tab 4 \tab 1 \tab 2 \tab 15 \tab 3 \tab 5 \tab 22 \tab 10 \tab 11 \tab 1 \tab 1 \tab 14 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 32 \tab 1 \tab 41 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 18 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 4 \tab 1 \tab 3 \tab 6 \tab 2 \tab 2 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 7 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 7 \tab 5 \tab 19 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 6 \tab 2 \tab 1 \tab 7 \tab 18 \tab 1 \tab 8 \tab 1 \tab 2 \tab 3 \tab 10 \tab 7 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 16 \tab 24 \tab 1 \tab 11 \tab 10 \tab 5 \tab 10 \tab 19 \tab 1 \tab 28 \tab 3 \tab 2 \tab 1 \tab 36 \tab - \tab 2992 \cr
   Total \tab 1587 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 6 \tab 3 \tab 39 \tab 13 \tab 4 \tab 1 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 10 \tab 2 \tab 1 \tab 7 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 22 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 7 \tab 2 \tab 2 \tab 2 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 18 \tab 6 \tab 2 \tab 14 \tab 1 \tab 7 \tab 2 \tab 2 \tab 12 \tab 3 \tab 18 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab 27 \tab 3 \tab 61 \tab 3 \tab 2 \tab 2 \tab 1 \tab 5 \tab 65 \tab 10 \tab 2 \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab 3 \tab 23 \tab 4 \tab 6 \tab 1 \tab 2 \tab 1 \tab 13 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 6 \tab 4 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 29 \tab 8 \tab 1 \tab 7 \tab 3 \tab 12 \tab 2 \tab 1 \tab 1 \tab 6 \tab 13 \tab 1 \tab 1 \tab 10 \tab 2 \tab 6 \tab 3 \tab 2 \tab 7 \tab 3 \tab 6 \tab 48 \tab 27 \tab 5 \tab 4 \tab 1 \tab 2 \tab 15 \tab 3 \tab 5 \tab 22 \tab 10 \tab 11 \tab 1 \tab 1 \tab 14 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 32 \tab 1 \tab 41 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 18 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 4 \tab 1 \tab 3 \tab 6 \tab 2 \tab 2 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 7 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 7 \tab 5 \tab 19 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 6 \tab 2 \tab 1 \tab 7 \tab 18 \tab 1 \tab 8 \tab 1 \tab 2 \tab 3 \tab 10 \tab 7 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 16 \tab 24 \tab 1 \tab 11 \tab 10 \tab 5 \tab 10 \tab 19 \tab 1 \tab 28 \tab 3 \tab 2 \tab 1 \tab 36 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{fstspyr}},
\code{\link{indfirst}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpicked}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{malive1}},
\code{\link{marnum}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{neverwk}},
\code{\link{palive1}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind10}},
\code{\link{sbind80}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sbocc80}},
\code{\link{sboccindv}},
\code{\link{sboccstatus}},
\code{\link{sbocctag}},
\code{\link{sbpicked}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{spfam16}},
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre}},
\code{\link{sphedpre10}},
\code{\link{sphedpre105plus}},
\code{\link{sphedslf}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spsibs}},
\code{\link{spyrborn}},
\code{\link{wrkslfst}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{Family Mobility}
\concept{Occupation}
\keyword{variable}
