% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{occmobil}
\alias{occmobil}
\title{Status of job higher than father's?}
\source{
General Social Survey https://gss.norc.org
}
\description{
occmobil
}
\details{
Question Please think of your present job (or your last one if you don't have one now). If you compare this job with the job your father had when you were 16, would you say that the level or status of your job is (or was) . . .
}
\section{Values}{

\itemize{
\item \code{1} much higher than dad
\item \code{2} higher
\item \code{3} about equal
\item \code{4} lower
\item \code{5} much lower than dad
\item \code{6} r never had job
\item \code{7} no dad-never had job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4313/vshow}{GSS Data Explorer page for \code{occmobil}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
