% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_02.R
\name{occonet}
\alias{occonet}
\title{Occupation code based on o*net}
\source{
General Social Survey https://gss.norc.org
}
\description{
occonet
}
\details{
Question A. What kind of work do you (did you normally) do? That is, what (is/was) your job called?
B. What (do/did) you actually do in that job? Tell me, what are/were, some of your main duties?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010 \cr
}
}

\seealso{
Other Core: 
\code{\link{commute}},
\code{\link{found}},
\code{\link{indus07}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{prestg80}}
}
\concept{Core}
\keyword{variable}
