% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{occtrain}
\alias{occtrain}
\title{Gone thru specialized training}
\source{
General Social Survey https://gss.norc.org
}
\description{
occtrain
}
\details{
Question A. Have you gone through any specialized training in school to obtain a vocational certificate, diploma, degree, or license for this occupation?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2138/vshow}{GSS Data Explorer page for \code{occtrain}}.
}

\seealso{
Other Jobs: 
\code{\link{occyrs}},
\code{\link{trainsch}},
\code{\link{typeorg}}
}
\concept{Jobs}
\keyword{variable}
