% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{odds1}
\alias{odds1}
\title{Test of knowledge about probablity1}
\source{
General Social Survey https://gss.norc.org
}
\description{
odds1
}
\details{
Question Now, think about this situation. A doctor tells a couple that their genetic makeup means that they’ve got one in four chances of having a child with an inherited illness.
A. Does this mean that if their first child has the illness, the next three will not have the illness?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3332/vshow}{GSS Data Explorer page for \code{odds1}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
