% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{offsup}
\alias{offsup}
\title{Do you supervise the work of other employees}
\source{
General Social Survey https://gss.norc.org
}
\description{
offsup
}
\details{
Question As an official part of your main job, do you supervise the work of other employees or tell other employees what work to do?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1313/vshow}{GSS Data Explorer page for \code{offsup}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 605 \tab 334 \tab 571 \tab 7 \tab 1517 \cr
   Total \tab 605 \tab 334 \tab 571 \tab 7 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{totsup}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
