% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_582.R
\name{old1}
\alias{old1}
\title{Age of 1st person}
\source{
General Social Survey https://gss.norc.org
}
\description{
old1
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
3. Age of first person
}
\section{Values}{

\itemize{
\item \code{97} 97 or older
\item \code{98} adult, age unspecified
\item \code{99} child, age unspecified
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5334/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 79 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 85 \tab 87 \tab 88 \tab 89 \tab don't know \tab no answer \tab 86 \tab 90 \tab 92 \tab 93 \tab 91 \tab 97 or older \tab 17 \tab 94 \tab 95 \tab 96 \tab 10 \tab 12 \tab 15 \tab 8 \tab 9 \tab adult, age unspecified \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 7 \tab 8 \tab 16 \tab 16 \tab 15 \tab 28 \tab 39 \tab 29 \tab 31 \tab 38 \tab 33 \tab 29 \tab 27 \tab 34 \tab 33 \tab 26 \tab 26 \tab 16 \tab 23 \tab 36 \tab 23 \tab 26 \tab 31 \tab 26 \tab 28 \tab 26 \tab 30 \tab 32 \tab 18 \tab 24 \tab 26 \tab 31 \tab 30 \tab 28 \tab 33 \tab 31 \tab 23 \tab 26 \tab 27 \tab 24 \tab 28 \tab 15 \tab 25 \tab 17 \tab 31 \tab 20 \tab 23 \tab 29 \tab 22 \tab 23 \tab 22 \tab 28 \tab 15 \tab 8 \tab 13 \tab 12 \tab 9 \tab 9 \tab 13 \tab 11 \tab 8 \tab 9 \tab 12 \tab 8 \tab 5 \tab 4 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 755 \tab 4 \tab 5 \tab 4 \tab 6 \tab 16 \tab 16 \tab 12 \tab 14 \tab 20 \tab 21 \tab 16 \tab 19 \tab 15 \tab 21 \tab 16 \tab 17 \tab 15 \tab 15 \tab 12 \tab 15 \tab 18 \tab 12 \tab 15 \tab 12 \tab 9 \tab 12 \tab 8 \tab 10 \tab 14 \tab 7 \tab 9 \tab 9 \tab 11 \tab 15 \tab 15 \tab 11 \tab 8 \tab 9 \tab 13 \tab 9 \tab 13 \tab 10 \tab 16 \tab 17 \tab 7 \tab 11 \tab 9 \tab 6 \tab 15 \tab 13 \tab 7 \tab 15 \tab 5 \tab 3 \tab 8 \tab 9 \tab 6 \tab 10 \tab 5 \tab 7 \tab 8 \tab 5 \tab 2 \tab 4 \tab 4 \tab 2 \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 2 \tab 6 \tab 9 \tab 21 \tab 15 \tab 26 \tab 31 \tab 32 \tab 32 \tab 31 \tab 33 \tab 38 \tab 30 \tab 17 \tab 30 \tab 34 \tab 29 \tab 33 \tab 29 \tab 25 \tab 25 \tab 24 \tab 24 \tab 32 \tab 33 \tab 20 \tab 31 \tab 27 \tab 29 \tab 32 \tab 32 \tab 40 \tab 34 \tab 23 \tab 35 \tab 27 \tab 32 \tab 33 \tab 33 \tab 29 \tab 34 \tab 34 \tab 19 \tab 24 \tab 32 \tab 18 \tab 21 \tab 20 \tab 18 \tab 29 \tab 15 \tab 15 \tab 12 \tab 8 \tab 23 \tab 18 \tab 11 \tab 14 \tab 14 \tab 9 \tab 5 \tab 4 \tab 6 \tab 3 \tab 5 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 3 \tab 4 \tab 13 \tab 18 \tab 22 \tab 35 \tab 28 \tab 41 \tab 33 \tab 39 \tab 33 \tab 37 \tab 47 \tab 48 \tab 29 \tab 37 \tab 38 \tab 28 \tab 33 \tab 24 \tab 29 \tab 31 \tab 34 \tab 23 \tab 32 \tab 32 \tab 16 \tab 19 \tab 19 \tab 21 \tab 33 \tab 20 \tab 23 \tab 28 \tab 30 \tab 26 \tab 18 \tab 31 \tab 28 \tab 32 \tab 22 \tab 22 \tab 20 \tab 24 \tab 19 \tab 19 \tab 19 \tab 25 \tab 19 \tab 17 \tab 11 \tab 20 \tab 20 \tab 8 \tab 18 \tab 18 \tab 11 \tab 15 \tab 9 \tab 12 \tab 14 \tab 6 \tab 9 \tab 3 \tab 5 \tab 6 \tab 4 \tab 6 \tab 2 \tab - \tab 3 \tab - \tab 6 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 6 \tab 8 \tab 14 \tab 20 \tab 21 \tab 37 \tab 22 \tab 27 \tab 45 \tab 19 \tab 34 \tab 38 \tab 33 \tab 28 \tab 51 \tab 29 \tab 29 \tab 38 \tab 31 \tab 32 \tab 12 \tab 29 \tab 26 \tab 21 \tab 21 \tab 30 \tab 16 \tab 18 \tab 21 \tab 18 \tab 25 \tab 23 \tab 28 \tab 22 \tab 27 \tab 26 \tab 22 \tab 26 \tab 28 \tab 24 \tab 32 \tab 28 \tab 17 \tab 15 \tab 24 \tab 16 \tab 19 \tab 29 \tab 18 \tab 24 \tab 15 \tab 22 \tab 22 \tab 15 \tab 18 \tab 14 \tab 10 \tab 11 \tab 14 \tab 9 \tab 8 \tab 9 \tab 8 \tab 4 \tab 6 \tab 5 \tab 4 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 6 \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1 \tab 6 \tab 12 \tab 24 \tab 22 \tab 24 \tab 40 \tab 39 \tab 50 \tab 51 \tab 47 \tab 50 \tab 45 \tab 38 \tab 39 \tab 18 \tab 41 \tab 39 \tab 40 \tab 35 \tab 40 \tab 30 \tab 36 \tab 29 \tab 26 \tab 31 \tab 27 \tab 27 \tab 26 \tab 20 \tab 27 \tab 33 \tab 27 \tab 31 \tab 29 \tab 32 \tab 28 \tab 43 \tab 33 \tab 37 \tab 34 \tab 29 \tab 41 \tab 24 \tab 34 \tab 18 \tab 28 \tab 24 \tab 27 \tab 38 \tab 18 \tab 28 \tab 24 \tab 25 \tab 20 \tab 14 \tab 17 \tab 13 \tab 8 \tab 15 \tab 12 \tab 10 \tab 7 \tab 8 \tab 12 \tab 5 \tab 11 \tab 5 \tab 5 \tab 2 \tab 1 \tab - \tab 18 \tab 5 \tab - \tab 2 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 2 \tab 5 \tab 7 \tab 12 \tab 20 \tab 33 \tab 33 \tab 37 \tab 48 \tab 50 \tab 34 \tab 33 \tab 47 \tab 38 \tab 39 \tab 34 \tab 35 \tab 37 \tab 38 \tab 34 \tab 29 \tab 31 \tab 30 \tab 27 \tab 34 \tab 32 \tab 24 \tab 25 \tab 21 \tab 24 \tab 20 \tab 27 \tab 24 \tab 19 \tab 27 \tab 23 \tab 24 \tab 27 \tab 28 \tab 24 \tab 24 \tab 29 \tab 27 \tab 15 \tab 24 \tab 17 \tab 26 \tab 18 \tab 21 \tab 25 \tab 22 \tab 18 \tab 17 \tab 18 \tab 22 \tab 13 \tab 19 \tab 11 \tab 10 \tab 11 \tab 8 \tab 8 \tab 10 \tab 6 \tab 11 \tab 8 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 9 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab 9 \tab 12 \tab 17 \tab 28 \tab 31 \tab 31 \tab 32 \tab 46 \tab 28 \tab 33 \tab 31 \tab 26 \tab 31 \tab 41 \tab 26 \tab 38 \tab 41 \tab 29 \tab 48 \tab 36 \tab 31 \tab 27 \tab 26 \tab 23 \tab 32 \tab 18 \tab 20 \tab 25 \tab 24 \tab 17 \tab 20 \tab 27 \tab 18 \tab 20 \tab 26 \tab 21 \tab 24 \tab 19 \tab 20 \tab 24 \tab 21 \tab 26 \tab 24 \tab 21 \tab 17 \tab 27 \tab 17 \tab 18 \tab 22 \tab 12 \tab 21 \tab 17 \tab 17 \tab 10 \tab 13 \tab 15 \tab 16 \tab 9 \tab 8 \tab 3 \tab 5 \tab 10 \tab 7 \tab 6 \tab 6 \tab 5 \tab 5 \tab 4 \tab 3 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 3 \tab 10 \tab 8 \tab 16 \tab 28 \tab 19 \tab 33 \tab 36 \tab 33 \tab 31 \tab 35 \tab 27 \tab 32 \tab 35 \tab 38 \tab 33 \tab 33 \tab 34 \tab 29 \tab 45 \tab 28 \tab 27 \tab 37 \tab 23 \tab 29 \tab 30 \tab 22 \tab 29 \tab 18 \tab 18 \tab 22 \tab 26 \tab 37 \tab 22 \tab 18 \tab 23 \tab 20 \tab 20 \tab 19 \tab 35 \tab 24 \tab 24 \tab 23 \tab 23 \tab 29 \tab 22 \tab 19 \tab 26 \tab 18 \tab 24 \tab 23 \tab 17 \tab 21 \tab 12 \tab 14 \tab 15 \tab 13 \tab 12 \tab 16 \tab 15 \tab 10 \tab 8 \tab 7 \tab 11 \tab 7 \tab 3 \tab 11 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab 7 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab 5 \tab 6 \tab 18 \tab 15 \tab 13 \tab 27 \tab 37 \tab 28 \tab 42 \tab 33 \tab 41 \tab 32 \tab 33 \tab 25 \tab 36 \tab 36 \tab 30 \tab 30 \tab 24 \tab 34 \tab 36 \tab 36 \tab 21 \tab 31 \tab 34 \tab 21 \tab 30 \tab 21 \tab 28 \tab 14 \tab 25 \tab 22 \tab 14 \tab 16 \tab 18 \tab 21 \tab 20 \tab 27 \tab 19 \tab 19 \tab 23 \tab 21 \tab 25 \tab 18 \tab 28 \tab 28 \tab 25 \tab 17 \tab 18 \tab 19 \tab 21 \tab 19 \tab 11 \tab 19 \tab 15 \tab 11 \tab 12 \tab 19 \tab 11 \tab 12 \tab 7 \tab 13 \tab 10 \tab 6 \tab 8 \tab 3 \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab 17 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 3 \tab 7 \tab 13 \tab 22 \tab 28 \tab 36 \tab 27 \tab 38 \tab 28 \tab 33 \tab 41 \tab 45 \tab 45 \tab 41 \tab 35 \tab 51 \tab 43 \tab 41 \tab 41 \tab 42 \tab 38 \tab 46 \tab 43 \tab 31 \tab 36 \tab 34 \tab 45 \tab 38 \tab 31 \tab 32 \tab 26 \tab 28 \tab 28 \tab 28 \tab 24 \tab 30 \tab 32 \tab 22 \tab 30 \tab 21 \tab 27 \tab 22 \tab 24 \tab 20 \tab 20 \tab 20 \tab 24 \tab 33 \tab 16 \tab 33 \tab 18 \tab 22 \tab 22 \tab 15 \tab 16 \tab 18 \tab 24 \tab 16 \tab 11 \tab 16 \tab 12 \tab 7 \tab 9 \tab 7 \tab 7 \tab 5 \tab 4 \tab 6 \tab 6 \tab 4 \tab 2 \tab 1 \tab 19 \tab 3 \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 3 \tab 10 \tab 19 \tab 19 \tab 18 \tab 28 \tab 26 \tab 28 \tab 28 \tab 37 \tab 34 \tab 31 \tab 30 \tab 23 \tab 32 \tab 21 \tab 41 \tab 32 \tab 40 \tab 45 \tab 26 \tab 37 \tab 33 \tab 31 \tab 31 \tab 29 \tab 29 \tab 24 \tab 18 \tab 18 \tab 16 \tab 21 \tab 29 \tab 13 \tab 21 \tab 23 \tab 12 \tab 15 \tab 16 \tab 16 \tab 25 \tab 24 \tab 18 \tab 20 \tab 21 \tab 21 \tab 15 \tab 19 \tab 24 \tab 34 \tab 16 \tab 18 \tab 23 \tab 18 \tab 20 \tab 14 \tab 19 \tab 12 \tab 10 \tab 16 \tab 13 \tab 7 \tab 5 \tab 8 \tab 9 \tab 2 \tab 4 \tab 2 \tab 2 \tab 5 \tab 3 \tab 1 \tab 20 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 2 \tab 2 \tab 18 \tab 12 \tab 18 \tab 14 \tab 30 \tab 25 \tab 28 \tab 33 \tab 34 \tab 22 \tab 36 \tab 36 \tab 32 \tab 29 \tab 35 \tab 46 \tab 43 \tab 38 \tab 30 \tab 33 \tab 35 \tab 29 \tab 37 \tab 25 \tab 23 \tab 38 \tab 34 \tab 33 \tab 17 \tab 27 \tab 37 \tab 22 \tab 18 \tab 15 \tab 21 \tab 25 \tab 17 \tab 23 \tab 16 \tab 18 \tab 16 \tab 22 \tab 15 \tab 21 \tab 14 \tab 33 \tab 15 \tab 21 \tab 28 \tab 27 \tab 16 \tab 20 \tab 27 \tab 12 \tab 15 \tab 11 \tab 15 \tab 13 \tab 13 \tab 8 \tab 10 \tab 13 \tab 3 \tab 6 \tab 4 \tab 4 \tab 3 \tab 3 \tab 4 \tab - \tab 12 \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 2 \tab 5 \tab 9 \tab 16 \tab 15 \tab 21 \tab 21 \tab 12 \tab 19 \tab 28 \tab 32 \tab 30 \tab 26 \tab 34 \tab 32 \tab 34 \tab 42 \tab 36 \tab 33 \tab 33 \tab 28 \tab 26 \tab 31 \tab 25 \tab 31 \tab 33 \tab 31 \tab 22 \tab 28 \tab 24 \tab 29 \tab 12 \tab 20 \tab 11 \tab 23 \tab 13 \tab 15 \tab 19 \tab 21 \tab 21 \tab 10 \tab 12 \tab 28 \tab 18 \tab 19 \tab 20 \tab 18 \tab 26 \tab 20 \tab 19 \tab 16 \tab 15 \tab 8 \tab 16 \tab 16 \tab 19 \tab 11 \tab 17 \tab 19 \tab 13 \tab 10 \tab 10 \tab 8 \tab 7 \tab 9 \tab 9 \tab 6 \tab 1 \tab 2 \tab 5 \tab 4 \tab 3 \tab 7 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 3 \tab 8 \tab 12 \tab 24 \tab 21 \tab 19 \tab 28 \tab 23 \tab 25 \tab 22 \tab 32 \tab 25 \tab 29 \tab 40 \tab 44 \tab 33 \tab 42 \tab 33 \tab 28 \tab 35 \tab 41 \tab 39 \tab 43 \tab 32 \tab 42 \tab 32 \tab 24 \tab 23 \tab 36 \tab 23 \tab 24 \tab 19 \tab 15 \tab 22 \tab 29 \tab 14 \tab 23 \tab 19 \tab 17 \tab 13 \tab 13 \tab 17 \tab 29 \tab 20 \tab 22 \tab 12 \tab 20 \tab 15 \tab 14 \tab 21 \tab 22 \tab 20 \tab 19 \tab 17 \tab 27 \tab 14 \tab 24 \tab 10 \tab 9 \tab 10 \tab 11 \tab 15 \tab 10 \tab 13 \tab 6 \tab 6 \tab 5 \tab 6 \tab 4 \tab 1 \tab 2 \tab - \tab 16 \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 4 \tab 5 \tab 6 \tab 12 \tab 16 \tab 17 \tab 23 \tab 22 \tab 24 \tab 36 \tab 31 \tab 27 \tab 34 \tab 33 \tab 29 \tab 44 \tab 34 \tab 41 \tab 41 \tab 34 \tab 34 \tab 48 \tab 44 \tab 33 \tab 42 \tab 39 \tab 24 \tab 46 \tab 41 \tab 23 \tab 40 \tab 28 \tab 38 \tab 26 \tab 23 \tab 26 \tab 18 \tab 20 \tab 21 \tab 15 \tab 22 \tab 20 \tab 17 \tab 14 \tab 27 \tab 19 \tab 19 \tab 21 \tab 12 \tab 14 \tab 18 \tab 16 \tab 23 \tab 23 \tab 25 \tab 18 \tab 20 \tab 11 \tab 19 \tab 12 \tab 11 \tab 6 \tab 9 \tab 12 \tab 9 \tab 7 \tab 4 \tab 4 \tab 2 \tab 4 \tab 3 \tab - \tab 13 \tab 6 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 4 \tab 11 \tab 10 \tab 22 \tab 30 \tab 33 \tab 45 \tab 47 \tab 37 \tab 58 \tab 61 \tab 65 \tab 61 \tab 73 \tab 78 \tab 81 \tab 75 \tab 76 \tab 71 \tab 81 \tab 75 \tab 57 \tab 81 \tab 61 \tab 61 \tab 68 \tab 66 \tab 57 \tab 71 \tab 59 \tab 72 \tab 56 \tab 45 \tab 50 \tab 53 \tab 45 \tab 46 \tab 48 \tab 42 \tab 38 \tab 40 \tab 33 \tab 34 \tab 25 \tab 38 \tab 36 \tab 22 \tab 46 \tab 28 \tab 37 \tab 28 \tab 30 \tab 32 \tab 30 \tab 30 \tab 31 \tab 35 \tab 37 \tab 23 \tab 29 \tab 22 \tab 22 \tab 15 \tab 18 \tab 15 \tab 12 \tab 7 \tab 11 \tab 6 \tab 7 \tab 5 \tab 1 \tab 14 \tab 8 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 4 \tab 17 \tab 16 \tab 31 \tab 34 \tab 43 \tab 66 \tab 45 \tab 55 \tab 51 \tab 56 \tab 75 \tab 50 \tab 65 \tab 66 \tab 60 \tab 60 \tab 70 \tab 72 \tab 92 \tab 65 \tab 68 \tab 76 \tab 63 \tab 77 \tab 57 \tab 59 \tab 53 \tab 67 \tab 74 \tab 64 \tab 59 \tab 62 \tab 52 \tab 56 \tab 41 \tab 40 \tab 40 \tab 33 \tab 37 \tab 32 \tab 28 \tab 31 \tab 35 \tab 34 \tab 21 \tab 25 \tab 41 \tab 24 \tab 29 \tab 33 \tab 29 \tab 23 \tab 20 \tab 23 \tab 22 \tab 27 \tab 25 \tab 25 \tab 20 \tab 21 \tab 22 \tab 13 \tab 10 \tab 13 \tab 10 \tab 6 \tab 7 \tab 7 \tab 8 \tab 7 \tab 1 \tab 42 \tab 9 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 3 \tab 14 \tab 22 \tab 22 \tab 23 \tab 35 \tab 48 \tab 39 \tab 40 \tab 75 \tab 55 \tab 50 \tab 49 \tab 54 \tab 57 \tab 71 \tab 78 \tab 70 \tab 68 \tab 67 \tab 84 \tab 93 \tab 81 \tab 64 \tab 60 \tab 61 \tab 65 \tab 58 \tab 60 \tab 50 \tab 55 \tab 50 \tab 64 \tab 40 \tab 54 \tab 43 \tab 56 \tab 45 \tab 41 \tab 38 \tab 28 \tab 27 \tab 27 \tab 29 \tab 36 \tab 26 \tab 32 \tab 33 \tab 25 \tab 26 \tab 25 \tab 28 \tab 34 \tab 26 \tab 31 \tab 31 \tab 33 \tab 19 \tab 32 \tab 11 \tab 20 \tab 18 \tab 26 \tab 8 \tab 14 \tab 10 \tab 13 \tab 11 \tab 9 \tab 3 \tab 1 \tab 1 \tab 19 \tab 7 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 9 \tab 16 \tab 22 \tab 30 \tab 32 \tab 36 \tab 41 \tab 65 \tab 47 \tab 49 \tab 58 \tab 44 \tab 59 \tab 63 \tab 55 \tab 46 \tab 60 \tab 49 \tab 64 \tab 77 \tab 70 \tab 66 \tab 64 \tab 64 \tab 70 \tab 69 \tab 59 \tab 69 \tab 61 \tab 64 \tab 73 \tab 52 \tab 60 \tab 44 \tab 64 \tab 46 \tab 37 \tab 51 \tab 32 \tab 31 \tab 33 \tab 45 \tab 39 \tab 36 \tab 32 \tab 17 \tab 20 \tab 36 \tab 29 \tab 26 \tab 21 \tab 23 \tab 33 \tab 23 \tab 30 \tab 26 \tab 21 \tab 17 \tab 33 \tab 24 \tab 21 \tab 21 \tab 16 \tab 15 \tab 17 \tab 17 \tab 20 \tab 16 \tab 6 \tab 3 \tab 6 \tab 1 \tab 12 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 7 \tab 11 \tab 22 \tab 27 \tab 35 \tab 38 \tab 46 \tab 44 \tab 48 \tab 40 \tab 48 \tab 50 \tab 55 \tab 49 \tab 71 \tab 63 \tab 51 \tab 63 \tab 63 \tab 45 \tab 43 \tab 63 \tab 64 \tab 55 \tab 59 \tab 40 \tab 62 \tab 72 \tab 58 \tab 59 \tab 53 \tab 49 \tab 54 \tab 59 \tab 53 \tab 44 \tab 43 \tab 70 \tab 33 \tab 30 \tab 42 \tab 38 \tab 38 \tab 34 \tab 39 \tab 35 \tab 35 \tab 27 \tab 31 \tab 24 \tab 28 \tab 34 \tab 36 \tab 33 \tab 25 \tab 22 \tab 30 \tab 34 \tab 19 \tab 19 \tab 15 \tab 16 \tab 19 \tab 11 \tab 16 \tab 16 \tab 13 \tab 12 \tab 6 \tab 5 \tab 6 \tab 4 \tab 42 \tab 13 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 7 \tab 12 \tab 18 \tab 25 \tab 23 \tab 39 \tab 34 \tab 45 \tab 59 \tab 43 \tab 43 \tab 38 \tab 59 \tab 49 \tab 56 \tab 56 \tab 65 \tab 62 \tab 56 \tab 50 \tab 60 \tab 60 \tab 74 \tab 52 \tab 52 \tab 64 \tab 75 \tab 51 \tab 52 \tab 70 \tab 63 \tab 70 \tab 65 \tab 54 \tab 68 \tab 43 \tab 47 \tab 57 \tab 49 \tab 57 \tab 37 \tab 40 \tab 47 \tab 46 \tab 38 \tab 27 \tab 33 \tab 32 \tab 38 \tab 37 \tab 29 \tab 24 \tab 29 \tab 14 \tab 23 \tab 14 \tab 24 \tab 22 \tab 16 \tab 15 \tab 17 \tab 18 \tab 12 \tab 20 \tab 9 \tab 11 \tab 10 \tab 15 \tab 9 \tab 8 \tab 3 \tab 11 \tab 46 \tab 4 \tab 5 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 8 \tab 13 \tab 25 \tab 36 \tab 34 \tab 47 \tab 63 \tab 76 \tab 68 \tab 75 \tab 66 \tab 64 \tab 67 \tab 68 \tab 76 \tab 78 \tab 98 \tab 100 \tab 98 \tab 86 \tab 100 \tab 72 \tab 105 \tab 101 \tab 95 \tab 88 \tab 94 \tab 99 \tab 113 \tab 96 \tab 95 \tab 101 \tab 122 \tab 66 \tab 95 \tab 98 \tab 94 \tab 85 \tab 82 \tab 64 \tab 83 \tab 78 \tab 79 \tab 47 \tab 66 \tab 66 \tab 50 \tab 53 \tab 47 \tab 40 \tab 48 \tab 34 \tab 38 \tab 44 \tab 43 \tab 40 \tab 39 \tab 31 \tab 27 \tab 31 \tab 31 \tab 30 \tab 25 \tab 26 \tab 28 \tab 15 \tab 18 \tab 17 \tab 8 \tab 9 \tab 8 \tab 82 \tab 80 \tab 15 \tab 8 \tab 4 \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 3 \tab 8 \tab 9 \tab 16 \tab 17 \tab 33 \tab 25 \tab 27 \tab 31 \tab 34 \tab 27 \tab 30 \tab 40 \tab 20 \tab 27 \tab 32 \tab 36 \tab 35 \tab 40 \tab 47 \tab 30 \tab 33 \tab 36 \tab 30 \tab 35 \tab 43 \tab 38 \tab 49 \tab 51 \tab 40 \tab 38 \tab 52 \tab 45 \tab 36 \tab 38 \tab 38 \tab 40 \tab 39 \tab 40 \tab 34 \tab 22 \tab 30 \tab 33 \tab 54 \tab 23 \tab 24 \tab 29 \tab 36 \tab 20 \tab 18 \tab 18 \tab 21 \tab 26 \tab 21 \tab 20 \tab 15 \tab 19 \tab 16 \tab 11 \tab 15 \tab 17 \tab 19 \tab 12 \tab 15 \tab 8 \tab 11 \tab 6 \tab 5 \tab 4 \tab 2 \tab 2 \tab - \tab 106 \tab 8 \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 3 \tab 9 \tab 14 \tab 19 \tab 16 \tab 27 \tab 29 \tab 39 \tab 38 \tab 34 \tab 33 \tab 34 \tab 48 \tab 32 \tab 34 \tab 27 \tab 35 \tab 37 \tab 33 \tab 34 \tab 29 \tab 32 \tab 44 \tab 41 \tab 37 \tab 46 \tab 35 \tab 40 \tab 43 \tab 34 \tab 30 \tab 47 \tab 48 \tab 33 \tab 42 \tab 41 \tab 31 \tab 51 \tab 35 \tab 23 \tab 36 \tab 37 \tab 37 \tab 39 \tab 34 \tab 37 \tab 23 \tab 26 \tab 31 \tab 35 \tab 27 \tab 24 \tab 22 \tab 17 \tab 13 \tab 18 \tab 18 \tab 17 \tab 10 \tab 12 \tab 16 \tab 10 \tab 12 \tab 11 \tab 13 \tab 14 \tab 8 \tab 9 \tab 7 \tab 6 \tab 7 \tab - \tab 64 \tab 4 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 3 \tab 3 \tab 16 \tab 12 \tab 15 \tab 19 \tab 23 \tab 24 \tab 33 \tab 24 \tab 29 \tab 33 \tab 43 \tab 30 \tab 40 \tab 33 \tab 39 \tab 39 \tab 28 \tab 37 \tab 40 \tab 26 \tab 42 \tab 36 \tab 41 \tab 40 \tab 24 \tab 49 \tab 39 \tab 50 \tab 42 \tab 40 \tab 42 \tab 44 \tab 39 \tab 41 \tab 34 \tab 41 \tab 41 \tab 25 \tab 36 \tab 31 \tab 41 \tab 30 \tab 35 \tab 31 \tab 32 \tab 37 \tab 25 \tab 20 \tab 19 \tab 26 \tab 20 \tab 23 \tab 17 \tab 20 \tab 17 \tab 14 \tab 11 \tab 17 \tab 23 \tab 11 \tab 15 \tab 16 \tab 12 \tab 7 \tab 9 \tab 11 \tab 10 \tab 4 \tab 7 \tab 2 \tab 23 \tab 11 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1 \tab 4 \tab 7 \tab 11 \tab 16 \tab 21 \tab 31 \tab 28 \tab 32 \tab 29 \tab 38 \tab 50 \tab 44 \tab 46 \tab 38 \tab 40 \tab 37 \tab 38 \tab 48 \tab 47 \tab 45 \tab 37 \tab 53 \tab 43 \tab 54 \tab 48 \tab 27 \tab 38 \tab 42 \tab 50 \tab 39 \tab 44 \tab 66 \tab 44 \tab 74 \tab 43 \tab 51 \tab 71 \tab 59 \tab 34 \tab 47 \tab 62 \tab 41 \tab 45 \tab 40 \tab 29 \tab 41 \tab 36 \tab 37 \tab 39 \tab 22 \tab 36 \tab 31 \tab 32 \tab 22 \tab 23 \tab 18 \tab 22 \tab 24 \tab 23 \tab 24 \tab 21 \tab 16 \tab 13 \tab 11 \tab 6 \tab 12 \tab 13 \tab 12 \tab 4 \tab 7 \tab 83 \tab 55 \tab 6 \tab 3 \tab 5 \tab 3 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2 \tab 8 \tab 15 \tab 23 \tab 31 \tab 34 \tab 30 \tab 36 \tab 48 \tab 35 \tab 60 \tab 48 \tab 41 \tab 31 \tab 44 \tab 43 \tab 61 \tab 52 \tab 41 \tab 47 \tab 43 \tab 47 \tab 43 \tab 32 \tab 41 \tab 56 \tab 46 \tab 41 \tab 37 \tab 52 \tab 45 \tab 45 \tab 72 \tab 57 \tab 72 \tab 61 \tab 49 \tab 56 \tab 64 \tab 65 \tab 60 \tab 70 \tab 48 \tab 48 \tab 45 \tab 45 \tab 47 \tab 48 \tab 43 \tab 51 \tab 49 \tab 33 \tab 36 \tab 25 \tab 40 \tab 26 \tab 21 \tab 29 \tab 28 \tab 23 \tab 20 \tab 24 \tab 17 \tab 18 \tab 15 \tab 22 \tab 11 \tab 7 \tab 7 \tab 4 \tab 7 \tab 50 \tab 73 \tab 13 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 7 \tab 3 \tab 11 \tab 19 \tab 18 \tab 22 \tab 28 \tab 28 \tab 26 \tab 38 \tab 34 \tab 40 \tab 31 \tab 34 \tab 32 \tab 49 \tab 40 \tab 47 \tab 41 \tab 34 \tab 44 \tab 47 \tab 51 \tab 36 \tab 33 \tab 30 \tab 43 \tab 36 \tab 38 \tab 38 \tab 44 \tab 35 \tab 48 \tab 36 \tab 54 \tab 52 \tab 49 \tab 46 \tab 49 \tab 48 \tab 52 \tab 31 \tab 56 \tab 34 \tab 33 \tab 30 \tab 41 \tab 37 \tab 42 \tab 33 \tab 38 \tab 29 \tab 50 \tab 31 \tab 19 \tab 24 \tab 26 \tab 25 \tab 24 \tab 16 \tab 18 \tab 6 \tab 13 \tab 14 \tab 10 \tab 7 \tab 12 \tab 13 \tab 4 \tab 9 \tab 4 \tab 62 \tab 38 \tab 6 \tab 5 \tab 3 \tab 3 \tab 8 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 208 \tab 2 \tab 9 \tab 16 \tab 22 \tab 23 \tab 20 \tab 28 \tab 41 \tab 43 \tab 55 \tab 52 \tab 47 \tab 62 \tab 57 \tab 67 \tab 72 \tab 52 \tab 64 \tab 65 \tab 68 \tab 52 \tab 57 \tab 77 \tab 73 \tab 63 \tab 51 \tab 55 \tab 63 \tab 61 \tab 42 \tab 61 \tab 63 \tab 71 \tab 60 \tab 65 \tab 62 \tab 62 \tab 68 \tab 82 \tab 76 \tab 75 \tab 75 \tab 76 \tab 86 \tab 66 \tab 81 \tab 82 \tab 59 \tab 89 \tab 83 \tab 65 \tab 79 \tab 58 \tab 62 \tab 52 \tab 61 \tab 68 \tab 32 \tab 43 \tab 35 \tab 45 \tab 29 \tab 38 \tab 16 \tab 16 \tab 24 \tab 18 \tab 14 \tab 6 \tab 5 \tab 6 \tab - \tab - \tab 13 \tab 10 \tab 5 \tab 3 \tab 3 \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 94 \tab 10 \tab - \tab 4032 \cr
   2022 \tab 1779 \tab 7 \tab 4 \tab 15 \tab 13 \tab 17 \tab 15 \tab 20 \tab 20 \tab 26 \tab 35 \tab 29 \tab 26 \tab 26 \tab 28 \tab 38 \tab 35 \tab 29 \tab 28 \tab 36 \tab 17 \tab 33 \tab 42 \tab 31 \tab 26 \tab 33 \tab 27 \tab 29 \tab 31 \tab 22 \tab 31 \tab 21 \tab 18 \tab 26 \tab 27 \tab 24 \tab 24 \tab 24 \tab 23 \tab 31 \tab 20 \tab 26 \tab 28 \tab 47 \tab 29 \tab 44 \tab 28 \tab 26 \tab 38 \tab 35 \tab 20 \tab 46 \tab 34 \tab 36 \tab 24 \tab 23 \tab 13 \tab 26 \tab 22 \tab 24 \tab 15 \tab 13 \tab 16 \tab 10 \tab 8 \tab 12 \tab 4 \tab 10 \tab 10 \tab 5 \tab 3 \tab 4 \tab - \tab 2 \tab 8 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 56 \tab 7 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 7343 \tab 115 \tab 245 \tab 416 \tab 601 \tab 682 \tab 859 \tab 1008 \tab 1076 \tab 1149 \tab 1214 \tab 1224 \tab 1217 \tab 1269 \tab 1229 \tab 1324 \tab 1288 \tab 1373 \tab 1370 \tab 1344 \tab 1374 \tab 1284 \tab 1306 \tab 1447 \tab 1198 \tab 1296 \tab 1259 \tab 1166 \tab 1234 \tab 1215 \tab 1178 \tab 1172 \tab 1170 \tab 1320 \tab 1044 \tab 1239 \tab 1086 \tab 1041 \tab 1170 \tab 1090 \tab 982 \tab 1016 \tab 1001 \tab 1041 \tab 939 \tab 966 \tab 832 \tab 867 \tab 951 \tab 838 \tab 893 \tab 778 \tab 807 \tab 787 \tab 659 \tab 707 \tab 622 \tab 657 \tab 562 \tab 550 \tab 493 \tab 481 \tab 407 \tab 394 \tab 343 \tab 319 \tab 271 \tab 249 \tab 229 \tab 149 \tab 117 \tab 108 \tab 312 \tab 779 \tab 189 \tab 79 \tab 48 \tab 29 \tab 69 \tab 16 \tab 1 \tab 17 \tab 16 \tab 13 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 150 \tab 17 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away1}},
\code{\link{away2}},
\code{\link{gender1}},
\code{\link{gender2}},
\code{\link{mar1}},
\code{\link{mar2}},
\code{\link{old2}},
\code{\link{relate2}},
\code{\link{where1}}

Other Family: 
\code{\link{away1}},
\code{\link{away2}},
\code{\link{gender1}},
\code{\link{gender2}},
\code{\link{mar1}},
\code{\link{mar2}},
\code{\link{old2}},
\code{\link{relate2}},
\code{\link{where1}}

Other Household: 
\code{\link{away1}},
\code{\link{away2}},
\code{\link{gender1}},
\code{\link{gender2}},
\code{\link{mar1}},
\code{\link{mar2}},
\code{\link{old2}},
\code{\link{relate2}},
\code{\link{where1}}
}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
