% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_585.R
\name{old6}
\alias{old6}
\title{Age of 6th person}
\source{
General Social Survey https://gss.norc.org
}
\description{
old6
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
3. Age of sixth person
}
\section{Values}{

\itemize{
\item \code{97} 97 or older
\item \code{98} adult, age unspecified
\item \code{99} child, age unspecified
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5364/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 22 \tab 24 \tab 29 \tab 3 \tab 4 \tab 5 \tab 53 \tab 56 \tab 58 \tab 6 \tab 7 \tab 73 \tab 8 \tab 9 \tab 67 \tab no answer \tab 20 \tab 21 \tab 26 \tab 71 \tab 23 \tab 65 \tab 36 \tab 48 \tab 32 \tab 74 \tab 25 \tab 35 \tab 39 \tab 75 \tab 41 \tab 78 \tab 37 \tab 81 \tab 28 \tab 55 \tab 27 \tab 60 \tab 45 \tab 49 \tab 40 \tab 69 \tab 50 \tab 51 \tab 64 \tab 33 \tab 83 \tab don't know \tab 63 \tab 72 \tab 77 \tab 84 \tab 30 \tab 31 \tab 34 \tab 86 \tab 79 \tab 80 \tab 59 \tab adult, age unspecified \tab child, age unspecified \tab 52 \tab 62 \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1363 \tab 2 \tab 2 \tab 8 \tab 9 \tab 10 \tab 14 \tab 9 \tab 7 \tab 1 \tab 4 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab 5 \tab 8 \tab 1 \tab 1 \tab 1 \tab 7 \tab 6 \tab 1 \tab 5 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1452 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab 2 \tab 8 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1413 \tab 4 \tab 7 \tab 3 \tab 8 \tab 11 \tab 6 \tab 6 \tab 3 \tab 4 \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab 5 \tab 8 \tab - \tab - \tab 1 \tab 5 \tab 6 \tab - \tab 6 \tab 14 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1424 \tab 8 \tab 8 \tab 3 \tab 7 \tab 4 \tab 7 \tab 8 \tab 5 \tab 4 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 8 \tab 5 \tab - \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1392 \tab 5 \tab 4 \tab 7 \tab 3 \tab 6 \tab 8 \tab 5 \tab - \tab 4 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 6 \tab 5 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1763 \tab 3 \tab 6 \tab 5 \tab 7 \tab 5 \tab 7 \tab 6 \tab 6 \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab 4 \tab 5 \tab 4 \tab - \tab - \tab - \tab 4 \tab 4 \tab - \tab 9 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1513 \tab 2 \tab 3 \tab 4 \tab 2 \tab 6 \tab 4 \tab 2 \tab 4 \tab 3 \tab 2 \tab 3 \tab - \tab 5 \tab - \tab 1 \tab - \tab 7 \tab 5 \tab 6 \tab - \tab - \tab - \tab 4 \tab 5 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1402 \tab 4 \tab 3 \tab 4 \tab 3 \tab 4 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 7 \tab 9 \tab 3 \tab - \tab - \tab - \tab 5 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1472 \tab 4 \tab 2 \tab 5 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 6 \tab 5 \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1402 \tab 7 \tab 3 \tab 2 \tab 5 \tab 1 \tab 3 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 7 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1727 \tab 7 \tab 10 \tab 4 \tab 2 \tab 3 \tab 6 \tab 2 \tab 3 \tab 3 \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 6 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1428 \tab 3 \tab 5 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 5 \tab 8 \tab 1 \tab - \tab - \tab - \tab 6 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1464 \tab 2 \tab 4 \tab 4 \tab 2 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 5 \tab 3 \tab 10 \tab - \tab - \tab - \tab 6 \tab 7 \tab - \tab 5 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1325 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab 7 \tab 3 \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1468 \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1566 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 5 \tab 3 \tab 2 \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2900 \tab 4 \tab 5 \tab 6 \tab 5 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab 9 \tab 9 \tab 7 \tab - \tab - \tab - \tab 5 \tab 5 \tab - \tab 5 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2791 \tab 5 \tab 7 \tab 6 \tab 3 \tab 4 \tab 7 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 10 \tab 2 \tab - \tab - \tab 8 \tab 9 \tab 7 \tab - \tab - \tab - \tab 5 \tab 5 \tab - \tab 4 \tab 10 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2741 \tab - \tab 13 \tab 8 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab 5 \tab - \tab - \tab - \tab 3 \tab 9 \tab 1 \tab 8 \tab 4 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2721 \tab 5 \tab 8 \tab 3 \tab 5 \tab 6 \tab 5 \tab 1 \tab 6 \tab 1 \tab 2 \tab - \tab 3 \tab 6 \tab 1 \tab 1 \tab - \tab 4 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 9 \tab - \tab 7 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2701 \tab - \tab 12 \tab 1 \tab 8 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 5 \tab 6 \tab 4 \tab - \tab - \tab - \tab 4 \tab 5 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2724 \tab - \tab 5 \tab 6 \tab 3 \tab 7 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 11 \tab - \tab - \tab - \tab 4 \tab 4 \tab 7 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab - \tab 7 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4358 \tab 6 \tab 16 \tab 8 \tab 6 \tab 8 \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 14 \tab 1 \tab 1 \tab - \tab 13 \tab 9 \tab 13 \tab - \tab - \tab - \tab 2 \tab 12 \tab - \tab 10 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1950 \tab 3 \tab 10 \tab 2 \tab 5 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 7 \tab - \tab - \tab - \tab 3 \tab 6 \tab 1 \tab - \tab - \tab - \tab 6 \tab 3 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1950 \tab 3 \tab 7 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab 2 \tab 11 \tab - \tab 1 \tab 1 \tab 4 \tab 5 \tab 4 \tab - \tab - \tab - \tab 4 \tab 8 \tab - \tab 4 \tab 9 \tab - \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1898 \tab 3 \tab 8 \tab 2 \tab 2 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 5 \tab 6 \tab 4 \tab - \tab - \tab - \tab 7 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2463 \tab 5 \tab 9 \tab 6 \tab 2 \tab 5 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab 3 \tab 4 \tab - \tab 6 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2770 \tab 3 \tab 10 \tab 2 \tab 2 \tab 2 \tab 4 \tab 4 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 6 \tab - \tab - \tab - \tab 4 \tab 6 \tab - \tab 9 \tab 4 \tab - \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2294 \tab - \tab 4 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 6 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 8 \tab - \tab 4 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3898 \tab 8 \tab 13 \tab 9 \tab 3 \tab 3 \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab 4 \tab 4 \tab - \tab - \tab 1 \tab 6 \tab 11 \tab 8 \tab - \tab - \tab - \tab 9 \tab 9 \tab 1 \tab 6 \tab 6 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3486 \tab 1 \tab 4 \tab 2 \tab 3 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 69820 \tab 104 \tab 203 \tab 124 \tab 114 \tab 112 \tab 118 \tab 86 \tab 69 \tab 53 \tab 53 \tab 31 \tab 45 \tab 158 \tab 20 \tab 11 \tab 6 \tab 164 \tab 165 \tab 145 \tab 3 \tab 4 \tab 2 \tab 129 \tab 150 \tab 3 \tab 144 \tab 124 \tab 2 \tab 48 \tab 22 \tab 21 \tab 7 \tab 2 \tab 16 \tab 3 \tab 2 \tab 4 \tab 4 \tab 1 \tab 9 \tab 7 \tab 2 \tab 6 \tab 1 \tab 2 \tab 4 \tab 2 \tab 4 \tab 2 \tab 5 \tab 4 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away6}},
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}

Other Family: 
\code{\link{away6}},
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}

Other Household: 
\code{\link{away6}},
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}
}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
