% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{onekid}
\alias{onekid}
\title{Desirability of one-kid family}
\source{
General Social Survey https://gss.norc.org
}
\description{
onekid
}
\details{
Question In general, what do you feel about each of these family sizes?
B. A family with one child
}
\section{Values}{

\itemize{
\item \verb{[1]} very desirable
\item \verb{[2]} desirable
\item \verb{[3]} neither
\item \verb{[4]} undesirable
\item \verb{[5]} very undesirable
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4376/vshow}{GSS Data Explorer page for \code{onekid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very desirable \tab desirable \tab neither \tab undesirable \tab very undesirable \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 96 \tab 583 \tab 307 \tab 239 \tab 33 \tab 59 \tab 97 \tab 1481 \cr
   Total \tab 67 \tab 96 \tab 583 \tab 307 \tab 239 \tab 33 \tab 59 \tab 97 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{mardiv}},
\code{\link{marhomo}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{threkids}},
\code{\link{twokids}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
