% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{onenatn}
\alias{onenatn}
\title{One nation or allowed to become fully separate?}
\source{
General Social Survey https://gss.norc.org
}
\description{
onenatn
}
\details{
Question Which of these two statements comes closer to your own view?
}
\section{Values}{

\itemize{
\item \code{1} essential to remain one nation
\item \code{2} allowed to become separate
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4826/vshow}{GSS Data Explorer page for \code{onenatn}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}}
}
\concept{ISSP National Identity}
\keyword{variable}
