% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_431.R
\name{ongorel2}
\alias{ongorel2}
\title{R had an ongoing relationship with leader2}
\source{
General Social Survey https://gss.norc.org
}
\description{
ongorel2
}
\details{
Question B. Did you become involved in an ongoing relationship with this leader?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3955/vshow}{GSS Data Explorer page for \code{ongorel2}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 2017 \tab 1 \tab 5 \tab 2023 \cr
   Total \tab 2017 \tab 1 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{hushrel1}},
\code{\link{ongorel1}},
\code{\link{ongorel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
