% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_467.R
\name{opable}
\alias{opable}
\title{Need natural ability to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opable
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
E. Natural ability? -- how important is that?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4245/vshow}{GSS Data Explorer page for \code{opable}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 210 \tab 738 \tab 521 \tab 53 \tab 7 \tab 15 \tab 20 \tab 1819 \cr
   Total \tab 255 \tab 210 \tab 738 \tab 521 \tab 53 \tab 7 \tab 15 \tab 20 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
