% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{opbribes}
\alias{opbribes}
\title{Need to give bribes to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opbribes
}
\details{
Question For each of these, please tell (CATIFILL3) how important you think it is for getting ahead in life ...how important is giving bribes?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7504/vshow}{GSS Data Explorer page for \code{opbribes}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 20 \tab 51 \tab 96 \tab 282 \tab 1269 \tab 148 \tab 120 \tab 17 \tab 4032 \cr
   Total \tab 2029 \tab 20 \tab 51 \tab 96 \tab 282 \tab 1269 \tab 148 \tab 120 \tab 17 \tab 4032 \cr
}
}

\keyword{variable}
