% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_467.R
\name{opclout}
\alias{opclout}
\title{Need political connections to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opclout
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
H. Having political connections?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1987 \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4248/vshow}{GSS Data Explorer page for \code{opclout}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 55 \tab 236 \tab 465 \tab 516 \tab 213 \tab 63 \tab 16 \tab - \tab 1819 \cr
   2021 \tab 2029 \tab 76 \tab 203 \tab 499 \tab 735 \tab 280 \tab 78 \tab 119 \tab 13 \tab 4032 \cr
   Total \tab 2284 \tab 131 \tab 439 \tab 964 \tab 1251 \tab 493 \tab 141 \tab 135 \tab 13 \tab 5851 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
