% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{opdevel}
\alias{opdevel}
\title{Opportunity to develop my abilities}
\source{
General Social Survey https://gss.norc.org
}
\description{
opdevel
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
B. I have an opportunity to develop my own special abilities
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not too true
\item \verb{[4]} not at all true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2797/vshow}{GSS Data Explorer page for \code{opdevel}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 743 \tab 686 \tab 220 \tab 121 \tab 11 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 661 \tab 689 \tab 226 \tab 120 \tab 19 \tab 19 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 463 \tab 475 \tab 142 \tab 75 \tab 7 \tab 25 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 479 \tab 516 \tab 180 \tab 66 \tab 2 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 521 \tab 625 \tab 194 \tab 69 \tab 8 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1552 \tab 723 \tab 827 \tab 257 \tab 134 \tab 15 \tab 15 \tab 21 \tab 3544 \cr
   Total \tab 8342 \tab 3590 \tab 3818 \tab 1219 \tab 585 \tab 62 \tab 112 \tab 21 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{cowrkint}},
\code{\link{fringeok}},
\code{\link{haveinfo}},
\code{\link{hlpequip}},
\code{\link{jobsecok}},
\code{\link{promtefr}},
\code{\link{supcares}},
\code{\link{wkfreedm}}
}
\concept{Quality of Working Life}
\keyword{variable}
