% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{opfamily}
\alias{opfamily}
\title{Family bkgd determines success in u.s.}
\source{
General Social Survey https://gss.norc.org
}
\description{
opfamily
}
\details{
Question Please indicate whether you agree or disagree with each of the following statements.
C. In America what you achieve in life depends largely on your family background.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4008/vshow}{GSS Data Explorer page for \code{opfamily}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
