% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_467.R
\name{ophrdwrk}
\alias{ophrdwrk}
\title{Need to work hard to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
ophrdwrk
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
F. Hard work -- how important is that for getting ahead in life?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1987 \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4246/vshow}{GSS Data Explorer page for \code{ophrdwrk}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 570 \tab 814 \tab 137 \tab 21 \tab 5 \tab 2 \tab 15 \tab - \tab 1819 \cr
   2021 \tab 2029 \tab 856 \tab 797 \tab 176 \tab 24 \tab 4 \tab 20 \tab 116 \tab 10 \tab 4032 \cr
   Total \tab 2284 \tab 1426 \tab 1611 \tab 313 \tab 45 \tab 9 \tab 22 \tab 131 \tab 10 \tab 5851 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
