% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_467.R
\name{opknow}
\alias{opknow}
\title{Need to know right people to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opknow
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
G. Knowing the right people?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4247/vshow}{GSS Data Explorer page for \code{opknow}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 137 \tab 548 \tab 654 \tab 173 \tab 27 \tab 9 \tab 16 \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 155 \tab 410 \tab 459 \tab 129 \tab 44 \tab 20 \tab 55 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 250 \tab 603 \tab 763 \tab 199 \tab 30 \tab 29 \tab 116 \tab 13 \tab 4032 \cr
   Total \tab 3829 \tab 542 \tab 1561 \tab 1876 \tab 501 \tab 101 \tab 58 \tab 187 \tab 13 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{oppared}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
