% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{oppsegov}
\alias{oppsegov}
\title{How important: citizen engage in acts of civil disobeyed}
\source{
General Social Survey https://gss.norc.org
}
\description{
oppsegov
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
F. That citizens may engage in acts of civil disobedience when they oppose government actions.
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4938/vshow}{GSS Data Explorer page for \code{oppsegov}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polgreed}},
\code{\link{polint1}}
}
\concept{Democracy}
\concept{ISSP Citizenship}
\keyword{variable}
