% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_467.R
\name{oprelig}
\alias{oprelig}
\title{Need to be right religion to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
oprelig
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
J. A person''s religion?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1987 \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4250/vshow}{GSS Data Explorer page for \code{oprelig}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 74 \tab 189 \tab 224 \tab 506 \tab 515 \tab 40 \tab 16 \tab - \tab 1819 \cr
   2021 \tab 2029 \tab 42 \tab 97 \tab 217 \tab 579 \tab 871 \tab 61 \tab 120 \tab 16 \tab 4032 \cr
   Total \tab 2284 \tab 116 \tab 286 \tab 441 \tab 1085 \tab 1386 \tab 101 \tab 136 \tab 16 \tab 5851 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oprace}},
\code{\link{opregion}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
