% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{oprich}
\alias{oprich}
\title{Child of rich has better opportunity}
\source{
General Social Survey https://gss.norc.org
}
\description{
oprich
}
\details{
Question Please indicate whether you agree or disagree with each of the following statements.
A. A person whose parents are rich has a better chance of earning a lot of money than a person whose parents are poor.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4006/vshow}{GSS Data Explorer page for \code{oprich}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab 5 \tab no answer \tab Total \cr
   1985 \tab 857 \tab 210 \tab 252 \tab 81 \tab 101 \tab 24 \tab 9 \tab 1534 \cr
   Total \tab 857 \tab 210 \tab 252 \tab 81 \tab 101 \tab 24 \tab 9 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{databank}},
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{progtax}},
\code{\link{verdict}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
