% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{orgfin}
\alias{orgfin}
\title{R's financial situation after org. changes}
\source{
General Social Survey https://gss.norc.org
}
\description{
orgfin
}
\details{
Question As a result of these organizational changes, would you say that your financial situation is now much better, better, worse, much worse, or have these things not affected your finances?
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} better
\item \code{3} worse
\item \code{4} much worse
\item \code{5} no effect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1270/vshow}{GSS Data Explorer page for \code{orgfin}}.
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{numunemp}},
\code{\link{totunemp}},
\code{\link{wrkyears}}
}
\concept{Work Organizations}
\keyword{variable}
