% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{orgsize}
\alias{orgsize}
\title{How many people employed in the entire org}
\source{
General Social Survey https://gss.norc.org
}
\description{
orgsize
}
\details{
Question D. About how many people are employed in the entire business, corporation, or organization for which you work? (Count part-time as well as full-time employees.)
}
\section{Values}{

\itemize{
\item \code{1} 1 - 9
\item \code{2} 10 - 49
\item \code{3} 50-99
\item \code{4} 100-499
\item \code{5} 500-999
\item \code{6} 1 000-1 999
\item \code{7} 2 000-9 999
\item \code{8} 10 000 or more
\item \code{98} not sure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2144/vshow}{GSS Data Explorer page for \code{orgsize}}.
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}

Other Job: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}
}
\concept{Job}
\concept{Jobs}
\keyword{variable}
