% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{orhealer}
\alias{orhealer}
\title{In what order go to a spiritual or a natural healer}
\source{
General Social Survey https://gss.norc.org
}
\description{
orhealer
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
F. Go to a spiritual or a natural healer for help.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2026/vshow}{GSS Data Explorer page for \code{orhealer}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2577 \tab 4 \tab 3 \tab 15 \tab 22 \tab 29 \tab 35 \tab 49 \tab 51 \tab 30 \tab 25 \tab 1 \tab 63 \tab 2904 \cr
   2006 \tab 3942 \tab 8 \tab 3 \tab 20 \tab 30 \tab 47 \tab 71 \tab 129 \tab 105 \tab 83 \tab 41 \tab - \tab 31 \tab 4510 \cr
   Total \tab 6519 \tab 12 \tab 6 \tab 35 \tab 52 \tab 76 \tab 106 \tab 178 \tab 156 \tab 113 \tab 66 \tab 1 \tab 94 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
