% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{ormntlhs}
\alias{ormntlhs}
\title{In what order check into a mental hospital}
\source{
General Social Survey https://gss.norc.org
}
\description{
ormntlhs
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
J. Check into a mental hospital.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2030/vshow}{GSS Data Explorer page for \code{ormntlhs}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2428 \tab 31 \tab 13 \tab 22 \tab 26 \tab 26 \tab 39 \tab 45 \tab 77 \tab 79 \tab 26 \tab 3 \tab 89 \tab 2904 \cr
   2006 \tab 4028 \tab 26 \tab 27 \tab 18 \tab 20 \tab 24 \tab 39 \tab 42 \tab 69 \tab 113 \tab 72 \tab - \tab 32 \tab 4510 \cr
   Total \tab 6456 \tab 57 \tab 40 \tab 40 \tab 46 \tab 50 \tab 78 \tab 87 \tab 146 \tab 192 \tab 98 \tab 3 \tab 121 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
