% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{ormntlot}
\alias{ormntlot}
\title{In what order go to a therapist, or counselor}
\source{
General Social Survey https://gss.norc.org
}
\description{
ormntlot
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
E. Go to a therapist, or counselor, like a psychologist, social worker, or other mental health professional for help.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2025/vshow}{GSS Data Explorer page for \code{ormntlot}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1706 \tab 124 \tab 161 \tab 223 \tab 232 \tab 184 \tab 84 \tab 32 \tab 10 \tab 2 \tab 3 \tab 143 \tab 2904 \cr
   2006 \tab 3272 \tab 57 \tab 157 \tab 236 \tab 294 \tab 280 \tab 113 \tab 48 \tab 9 \tab 4 \tab - \tab 40 \tab 4510 \cr
   Total \tab 4978 \tab 181 \tab 318 \tab 459 \tab 526 \tab 464 \tab 197 \tab 80 \tab 19 \tab 6 \tab 3 \tab 183 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
