% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{orslfhlp}
\alias{orslfhlp}
\title{In what order join a self-help group}
\source{
General Social Survey https://gss.norc.org
}
\description{
orslfhlp
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
G. Join a self-help group where people with similar problems help each other.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2027/vshow}{GSS Data Explorer page for \code{orslfhlp}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab 10 \tab Total \cr
   1996 \tab 1689 \tab 80 \tab 154 \tab 186 \tab 195 \tab 197 \tab 158 \tab 72 \tab 22 \tab 4 \tab 2 \tab 145 \tab - \tab 2904 \cr
   2006 \tab 3335 \tab 41 \tab 103 \tab 126 \tab 165 \tab 209 \tab 229 \tab 192 \tab 52 \tab 15 \tab - \tab 41 \tab 2 \tab 4510 \cr
   Total \tab 5024 \tab 121 \tab 257 \tab 312 \tab 360 \tab 406 \tab 387 \tab 264 \tab 74 \tab 19 \tab 2 \tab 186 \tab 2 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
