% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{ortlkclr}
\alias{ortlkclr}
\title{In what order talk to a minister, priest, rabbi etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
ortlkclr
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
B. Talk to a minister, priest, rabbi or other religious leader.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2022/vshow}{GSS Data Explorer page for \code{ortlkclr}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1699 \tab 109 \tab 1 \tab 316 \tab 183 \tab 154 \tab 122 \tab 90 \tab 61 \tab 22 \tab 4 \tab 3 \tab 140 \tab 2904 \cr
   2006 \tab 3288 \tab 74 \tab - \tab 379 \tab 185 \tab 131 \tab 121 \tab 124 \tab 100 \tab 60 \tab 13 \tab - \tab 35 \tab 4510 \cr
   Total \tab 4987 \tab 183 \tab 1 \tab 695 \tab 368 \tab 285 \tab 243 \tab 214 \tab 161 \tab 82 \tab 17 \tab 3 \tab 175 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
