% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{ortlkfm}
\alias{ortlkfm}
\title{In what order talk things over with family & friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
ortlkfm
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
A. Talk things over with family and friends.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2021/vshow}{GSS Data Explorer page for \code{ortlkfm}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1527 \tab 701 \tab 1 \tab 186 \tab 123 \tab 102 \tab 62 \tab 41 \tab 14 \tab 9 \tab 4 \tab 2 \tab 132 \tab 2904 \cr
   2006 \tab 3147 \tab 802 \tab 1 \tab 145 \tab 111 \tab 79 \tab 76 \tab 51 \tab 37 \tab 12 \tab 7 \tab - \tab 42 \tab 4510 \cr
   Total \tab 4674 \tab 1503 \tab 2 \tab 331 \tab 234 \tab 181 \tab 138 \tab 92 \tab 51 \tab 21 \tab 11 \tab 2 \tab 174 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}}
}
\concept{Mental Health}
\keyword{variable}
