% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{otcmed}
\alias{otcmed}
\title{X should take non-prescription medication}
\source{
General Social Survey https://gss.norc.org
}
\description{
otcmed
}
\details{
Question Should NAME do any of the following:
H. Take non-prescription medication, like over the counter sleeping pills.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2018/vshow}{GSS Data Explorer page for \code{otcmed}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1460 \tab 110 \tab 1250 \tab 52 \tab 32 \tab - \tab - \tab 2904 \cr
   2006 \tab 3073 \tab 178 \tab 1196 \tab 46 \tab 17 \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 184 \tab 930 \tab 30 \tab 29 \tab - \tab - \tab 2348 \cr
   2024 \tab 1693 \tab 323 \tab 1200 \tab 52 \tab 10 \tab - \tab 31 \tab 3309 \cr
   Total \tab 7401 \tab 795 \tab 4576 \tab 180 \tab 88 \tab 0 \tab 31 \tab 13071 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}

Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}
}
\concept{Mental Health}
\keyword{variable}
