% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_566.R
\name{oth1}
\alias{oth1}
\title{Being discriminated against}
\source{
General Social Survey https://gss.norc.org
}
\description{
oth1
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
H. And finally, since (CURRENT MONTH), (1990/2003), did any of the following occur to you?
\enumerate{
\item Discriminated against because of your race or nationality, sex, or for a similar reason.
}
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5189/vshow}{GSS Data Explorer page for \code{oth1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1991 \tab 500 \tab 60 \tab 953 \tab 4 \tab - \tab 1517 \cr
   2004 \tab 1472 \tab 69 \tab 1261 \tab 7 \tab 3 \tab 2812 \cr
   Total \tab 1972 \tab 129 \tab 2214 \tab 11 \tab 3 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}}
}
\concept{Negative Life Events}
\keyword{variable}
