% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_566.R
\name{oth3}
\alias{oth3}
\title{Child have trouble in school}
\source{
General Social Survey https://gss.norc.org
}
\description{
oth3
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
H. And finally, since (CURRENT MONTH), (1990/2003), did any of the following occur to you?
3. Child had serious trouble in school (e.g. failing, dropping out, expelled, held back).
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5191/vshow}{GSS Data Explorer page for \code{oth3}}.
}

\seealso{
Other Negative Life Events: 
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth4}}

Other Children: 
\code{\link{law7}}

Other Trauma: 
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth4}}
}
\concept{Children}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
