% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{oth5}
\alias{oth5}
\title{Do not have a car}
\source{
General Social Survey https://gss.norc.org
}
\description{
oth5
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
H. And finally, since (CURRENT MONTH), (1990/2003), did any of the following occur to you?
5. Did not have a car (one month or more).
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5193/vshow}{GSS Data Explorer page for \code{oth5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 500 \tab 131 \tab 882 \tab 4 \tab 1517 \cr
   2004 \tab 1472 \tab 125 \tab 1208 \tab 7 \tab 2812 \cr
   Total \tab 1972 \tab 256 \tab 2090 \tab 11 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{extra1}},
\code{\link{extra2}}
}
\concept{Negative Life Events}
\keyword{variable}
