% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{othbest}
\alias{othbest}
\title{What language other than english does r speak b}
\source{
General Social Survey https://gss.norc.org
}
\description{
othbest
}
\details{
Question D. Which of these languages other than English do you speak most fluently?
}
\section{Values}{

\itemize{
\item \verb{[1]} english
\item \verb{[2]} spanish
\item \verb{[3]} polish
\item \verb{[4]} french
\item \verb{[5]} hebrew
\item \verb{[6]} russian
\item \verb{[7]} tagalog
\item \verb{[8]} chinese
\item \verb{[9]} urdu
\item \verb{[10]} italian
\item \verb{[11]} romanian
\item \verb{[12]} german
\item \verb{[13]} pashtu
\item \verb{[14]} chamorro
\item \verb{[15]} ukranian
\item \verb{[16]} hungarian
\item \verb{[17]} turkish
\item \verb{[18]} khmer
\item \verb{[19]} korean
\item \verb{[20]} navajo
\item \verb{[21]} malayam
\item \verb{[22]} arabic
\item \verb{[23]} farsi-persian
\item \verb{[24]} serbian
\item \verb{[25]} lithuanian
\item \verb{[27]} thai
\item \verb{[28]} gujarati
\item \verb{[30]} dari
\item \verb{[31]} slovak
\item \verb{[32]} japanese
\item \verb{[33]} portuguese
\item \verb{[34]} vajun
\item \verb{[35]} yidddish
\item \verb{[36]} creole
\item \verb{[37]} vietnamese
\item \verb{[38]} norwegian
\item \verb{[39]} malaysian
\item \verb{[40]} latin
\item \verb{[41]} czech
\item \verb{[42]} hindu
\item \verb{[43]} hakka
\item \verb{[44]} swedish
\item \verb{[45]} vesaya
\item \verb{[46]} sign language
\item \verb{[47]} tongan
\item \verb{[48]} armenian
\item \verb{[49]} ibo
\item \verb{[50]} dutch
\item \verb{[51]} cherokee
\item \verb{[52]} samoan
\item \verb{[53]} yoruba
\item \verb{[54]} greek
\item \verb{[55]} lakota
\item \verb{[56]} african
\item \verb{[57]} anishnabee
\item \verb{[58]} ilokand
\item \verb{[59]} assyrian
\item \verb{[60]} danish
\item \verb{[61]} laotian
\item \verb{[62]} finnish
\item \verb{[63]} mohawk
\item \verb{[64]} amharic
\item \verb{[65]} bangladesh
\item \verb{[66]} oriya
\item \verb{[67]} hawaiian
\item \verb{[69]} hmong
\item \verb{[70]} indonesian
\item \verb{[71]} swahili
\item \verb{[72]} telugu
\item \verb{[73]} kannada
\item \verb{[74]} bulgarian
\item \verb{[75]} guarani
\item \verb{[76]} bosnian
\item \verb{[77]} punjabi
\item \verb{[78]} latavain
\item \verb{[79]} basque
\item \verb{[80]} bicol
\item \verb{[81]} croatian
\item \verb{[82]} tigan gypsy
\item \verb{[83]} slovenian
\item \verb{[84]} belgian
\item \verb{[85]} mandingo
\item \verb{[86]} jamaca
\item \verb{[87]} somali
\item \verb{[88]} native amerian
\item \verb{[89]} tagringa
\item \verb{[90]} krio
\item \verb{[91]} indian
\item \verb{[92]} cebuano
\item \verb{[93]} macedonian
\item \verb{[94]} patois
\item \verb{[95]} pangasinan
\item \verb{[96]} ilocano
\item \verb{[97]} marathi
\item \verb{[98]} kashmiri/cashmiri
\item \verb{[99]} tamil
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2286/vshow}{GSS Data Explorer page for \code{othbest}}.

Counts by year:\tabular{llllllllllllllllllllllllllllll}{
   year \tab iap \tab spanish \tab polish \tab french \tab hebrew \tab russian \tab chinese \tab italian \tab german \tab hungarian \tab arabic \tab farsi-persian \tab serbian \tab slovak \tab portuguese \tab creole \tab vietnamese \tab hindu \tab sign language \tab tongan \tab armenian \tab ibo \tab dutch \tab cherokee \tab patois \tab pangasinan \tab tamil \tab not available in this year \tab Total \cr
   2000 \tab 2728 \tab 28 \tab 1 \tab 15 \tab 1 \tab 1 \tab 3 \tab 4 \tab 7 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab 2817 \cr
   Total \tab 2728 \tab 28 \tab 1 \tab 15 \tab 1 \tab 1 \tab 3 \tab 4 \tab 7 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 0 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{betrlang}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blkwrk}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{ethchng}},
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{knwwht}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{leastcom}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{nobiling}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{uselang}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{uswht}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}

Other Language: 
\code{\link{betrlang}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{english}},
\code{\link{engoff1}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{uselang}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
