% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{othchrch}
\alias{othchrch}
\title{Happy in other non catholic church}
\source{
General Social Survey https://gss.norc.org
}
\description{
othchrch
}
\details{
Question Please tell me whether you strongly agree, agree, disagree, or strongly disagree with the following statements:
A. I could be just as happy in some other church, it wouldn't have to be Catholic.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1158/vshow}{GSS Data Explorer page for \code{othchrch}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2135 \tab 58 \tab 236 \tab 195 \tab 116 \tab 54 \tab 23 \tab 2817 \cr
   Total \tab 2135 \tab 58 \tab 236 \tab 195 \tab 116 \tab 54 \tab 23 \tab 2817 \cr
}
}

\seealso{
Other Religion: 
\code{\link{chrissch}},
\code{\link{denomsch}},
\code{\link{implives}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}}
}
\concept{Religion}
\keyword{variable}
