% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{othcredt}
\alias{othcredt}
\title{Other people take credit for r's work or ideas}
\source{
General Social Survey https://gss.norc.org
}
\description{
othcredt
}
\details{
Question C. Other people take credit for my work or ideas.
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3000/vshow}{GSS Data Explorer page for \code{othcredt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 117 \tab 395 \tab 469 \tab 820 \tab 23 \tab 43 \tab 2812 \cr
   2012 \tab 795 \tab 98 \tab 279 \tab 324 \tab 429 \tab 15 \tab 34 \tab 1974 \cr
   Total \tab 1740 \tab 215 \tab 674 \tab 793 \tab 1249 \tab 38 \tab 77 \tab 4786 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{difstand}},
\code{\link{lackinfo}},
\code{\link{perspace}},
\code{\link{putdown}},
\code{\link{rptprobs}}
}
\concept{Workplace Conflict}
\keyword{variable}
