% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_141.R
\name{othduty2}
\alias{othduty2}
\title{You train employees in other ways, not supervise}
\source{
General Social Survey https://gss.norc.org
}
\description{
othduty2
}
\details{
Question (On the card you have are/Here is a card listing) 7 different kinds of work activities. Which ones do you do with employees whom you do not officially supervise?  Please mention all that apply:
B. Train employees in other ways.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1326/vshow}{GSS Data Explorer page for \code{othduty2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 605 \tab 365 \tab 507 \tab 40 \tab 1517 \cr
   Total \tab 605 \tab 365 \tab 507 \tab 40 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{othduty1}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}}
}
\concept{Work Organizations}
\keyword{variable}
