% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_101.R
\name{other3}
\alias{other3}
\title{Person #3 is other type of acquaintance}
\source{
General Social Survey https://gss.norc.org
}
\description{
other3
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
K3. Third Person is Other
}
\section{Values}{

\itemize{
\item \verb{[1]} mentioned
\item \verb{[2]} not mentioned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/927/vshow}{GSS Data Explorer page for \code{other3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab don't know \tab Total \cr
   1985 \tab 599 \tab 11 \tab 920 \tab 4 \tab - \tab 1534 \cr
   1987 \tab 735 \tab 23 \tab 1036 \tab 25 \tab - \tab 1819 \cr
   2004 \tab 2240 \tab 10 \tab 514 \tab 47 \tab 1 \tab 2812 \cr
   Total \tab 3574 \tab 44 \tab 2470 \tab 76 \tab 1 \tab 6165 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other4}},
\code{\link{other5}}
}
\concept{Social Networks}
\keyword{variable}
