% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{otherwrk}
\alias{otherwrk}
\title{Does r have any additional job}
\source{
General Social Survey https://gss.norc.org
}
\description{
otherwrk
}
\details{
Question In addition to your main job, do you do any other work for additional income?
FI INSTRUCTION: PROBE 'YES' RESPONSES
}
\section{Values}{

\itemize{
\item \verb{[1]} no
\item \verb{[2]} yes, mostly as an employee
\item \verb{[3]} yes, mostly on a self-employed basis
\item \verb{[4]} yes, other
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4624/vshow}{GSS Data Explorer page for \code{otherwrk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab yes, mostly as an employee \tab yes, mostly on a self-employed basis \tab yes, other \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 764 \tab 71 \tab 151 \tab 28 \tab 1 \tab 3 \tab 4510 \cr
   Total \tab 3492 \tab 764 \tab 71 \tab 151 \tab 28 \tab 1 \tab 3 \tab 4510 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{skltrain}},
\code{\link{spwrkgvt}}
}
\concept{Work Orientation}
\keyword{variable}
