% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{othguns}
\alias{othguns}
\title{Number of adult owing a gun in household}
\source{
General Social Survey https://gss.norc.org
}
\description{
othguns
}
\details{
Question And how many other adults in your household personally own a gun?
}
\section{Values}{

\itemize{
\item \code{1} 1 person
\item \code{2} 2 persons
\item \code{3} 3 persons
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3705/vshow}{GSS Data Explorer page for \code{othguns}}.
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{rifles50}},
\code{\link{semiguns}}

Other Guns: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{rifles50}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\concept{Guns}
\keyword{variable}
