% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{othhome}
\alias{othhome}
\title{R speaks other language at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
othhome
}
\details{
Question A. Do you speak (LANGUAGE MENTIONED IN OTHLANG1) at home?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2301/vshow}{GSS Data Explorer page for \code{othhome}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2006 \tab 3837 \tab 258 \tab 413 \tab 2 \tab 4510 \cr
   Total \tab 3837 \tab 258 \tab 413 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{spklangw}},
\code{\link{uswht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
