% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{othhrs}
\alias{othhrs}
\title{Hours r spends at other job(s)}
\source{
General Social Survey https://gss.norc.org
}
\description{
othhrs
}
\details{
Question How many hours per week do you normally work in these other jobs, not including time spent on traveling to work?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4565/vshow}{GSS Data Explorer page for \code{othhrs}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 2 \tab 20 \tab 23 \tab 25 \tab 26 \tab 3 \tab 30 \tab 33 \tab 35 \tab 4 \tab 40 \tab 42 \tab 45 \tab 47 \tab 48 \tab 5 \tab 50 \tab 56 \tab 58 \tab 6 \tab 60 \tab 65 \tab 72 \tab 8 \tab 80 \tab 89 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1989 \tab 1283 \tab 8 \tab 25 \tab 9 \tab 2 \tab 3 \tab 19 \tab 3 \tab 16 \tab 16 \tab 1 \tab 5 \tab 1 \tab 14 \tab 7 \tab 1 \tab 3 \tab 12 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 4 \tab 1 \tab 1 \tab 10 \tab 3 \tab 1 \tab 1 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 43 \tab 1537 \cr
   Total \tab 1283 \tab 8 \tab 25 \tab 9 \tab 2 \tab 3 \tab 19 \tab 3 \tab 16 \tab 16 \tab 1 \tab 5 \tab 1 \tab 14 \tab 7 \tab 1 \tab 3 \tab 12 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 4 \tab 1 \tab 1 \tab 10 \tab 3 \tab 1 \tab 1 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 43 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{totalnum}}
}
\concept{Work Orientation}
\keyword{variable}
