% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{othjew16}
\alias{othjew16}
\title{R raised as a jew}
\source{
General Social Survey https://gss.norc.org
}
\description{
othjew16
}
\details{
Question Were you raised as a Jew?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} half/part
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/344/vshow}{GSS Data Explorer page for \code{othjew16}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab half/part \tab no \tab no answer \tab Total \cr
   2008 \tab 1837 \tab 1 \tab 173 \tab 12 \tab 2023 \cr
   2012 \tab 1710 \tab 1 \tab 259 \tab 4 \tab 1974 \cr
   Total \tab 3547 \tab 2 \tab 432 \tab 16 \tab 3997 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom16}},
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{oth16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}
}
\concept{Core}
\keyword{variable}
