% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{othkids}
\alias{othkids}
\title{R's current insurance plan covers children for whom r not a bio/adoptive parent}
\source{
General Social Survey https://gss.norc.org
}
\description{
othkids
}
\details{
Question Are the following people in your family covered by your current insurance plan?
C. Children for whom I am not a biological or adoptive parent
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} not applicable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3934/vshow}{GSS Data Explorer page for \code{othkids}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{dwell5}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evkid}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{mykids}},
\code{\link{negjob5}},
\code{\link{relgendr}},
\code{\link{sexsex18}},
\code{\link{sexunion}},
\code{\link{sppart}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
