% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{othlang}
\alias{othlang}
\title{Can r speak language other than english/spanish}
\source{
General Social Survey https://gss.norc.org
}
\description{
othlang
}
\details{
Question A. Can you speak a language other than English (Spanish)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2283/vshow}{GSS Data Explorer page for \code{othlang}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2000 \tab 1419 \tab 363 \tab 1031 \tab 4 \tab - \tab - \tab 2817 \cr
   2006 \tab 1702 \tab 671 \tab 2135 \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 555 \tab 1468 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 536 \tab 1508 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 574 \tab 1397 \tab 3 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 748 \tab 1789 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 816 \tab 2045 \tab 6 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 723 \tab 1623 \tab - \tab 2 \tab - \tab 2348 \cr
   2021 \tab - \tab 1144 \tab 2794 \tab 26 \tab - \tab 68 \tab 4032 \cr
   2022 \tab - \tab 979 \tab 2545 \tab 1 \tab - \tab 19 \tab 3544 \cr
   2024 \tab - \tab 904 \tab 2393 \tab - \tab - \tab 12 \tab 3309 \cr
   Total \tab 3121 \tab 8013 \tab 20728 \tab 43 \tab 2 \tab 99 \tab 32006 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{betrlang}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blkwrk}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{ethchng}},
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{knwwht}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{leastcom}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{uselang}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{uswht}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}

Other Language: 
\code{\link{betrlang}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{english}},
\code{\link{engoff1}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{uselang}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
