% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{othlang1}
\alias{othlang1}
\title{What other languages does r speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
othlang1
}
\details{
Question B. What other language(s) do you speak?  First response.
}
\section{Values}{

\itemize{
\item \verb{[1]} english
\item \verb{[2]} spanish
\item \verb{[3]} french
\item \verb{[4]} chinese
\item \verb{[5]} italian
\item \verb{[6]} german
\item \verb{[7]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}

Other Language: 
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
