% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_693.R
\name{othlang1_0022}
\alias{othlang1_0022}
\title{What other languages does r speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
othlang1_0022
}
\details{
Question What other language(s) do you speak?  First response.
As of 2024, OTHLANG1 has been recoded to only account for seven language categories. This has been retroactively applied to all previous years of OTHLANG1. The original OTHLANG1 containing data from 2000 to 2022 has been renamed OTHLANG1_0022.
}
\section{Values}{

\itemize{
\item \verb{[1]} english
\item \verb{[2]} spanish
\item \verb{[3]} polish
\item \verb{[4]} french
\item \verb{[5]} hebrew
\item \verb{[6]} russian
\item \verb{[7]} tagalog
\item \verb{[8]} chinese
\item \verb{[9]} urdu
\item \verb{[10]} italian
\item \verb{[11]} romanian
\item \verb{[12]} german
\item \verb{[13]} pashtu
\item \verb{[14]} chamorro
\item \verb{[15]} ukranian
\item \verb{[16]} hungarian
\item \verb{[17]} turkish
\item \verb{[18]} khmer
\item \verb{[19]} korean
\item \verb{[20]} navajo
\item \verb{[21]} malayam
\item \verb{[22]} arabic
\item \verb{[23]} farsi/persian
\item \verb{[24]} serbian
\item \verb{[25]} lithuanian
\item \verb{[27]} thai
\item \verb{[28]} gujarati
\item \verb{[30]} dari
\item \verb{[31]} slovak
\item \verb{[32]} japanese
\item \verb{[33]} portuguese
\item \verb{[34]} vajun
\item \verb{[35]} yiddish
\item \verb{[36]} creole
\item \verb{[37]} vietnamese
\item \verb{[38]} norwegian
\item \verb{[39]} malaysian
\item \verb{[40]} latin
\item \verb{[41]} czech
\item \verb{[42]} hindi
\item \verb{[43]} hakka
\item \verb{[44]} swedish
\item \verb{[45]} vesaya
\item \verb{[46]} sign language
\item \verb{[47]} tongan
\item \verb{[48]} armenian
\item \verb{[49]} ibo
\item \verb{[50]} dutch
\item \verb{[51]} cherokee
\item \verb{[52]} samoan
\item \verb{[53]} yoruba
\item \verb{[54]} greek
\item \verb{[55]} lakota
\item \verb{[56]} african
\item \verb{[57]} anishnabee
\item \verb{[58]} ilokand
\item \verb{[59]} assyrian
\item \verb{[60]} danish
\item \verb{[61]} laotian
\item \verb{[62]} finnish
\item \verb{[63]} mohawk
\item \verb{[64]} amharic
\item \verb{[65]} bengali
\item \verb{[66]} oriya
\item \verb{[67]} hawaiian
\item \verb{[68]} albanian
\item \verb{[69]} hmong
\item \verb{[70]} indonesian
\item \verb{[71]} swahili
\item \verb{[72]} telugu
\item \verb{[73]} kannada
\item \verb{[74]} bulgarian
\item \verb{[75]} guarani
\item \verb{[76]} bosnian
\item \verb{[77]} punjabi
\item \verb{[78]} latvian
\item \verb{[79]} basque
\item \verb{[80]} bikol/bicolano
\item \verb{[81]} croatian
\item \verb{[82]} romani
\item \verb{[83]} slovenian
\item \verb{[84]} belgian
\item \verb{[85]} mandingo
\item \verb{[86]} jamaican
\item \verb{[87]} somali
\item \verb{[88]} other native american language
\item \verb{[89]} tigrinya
\item \verb{[90]} krio
\item \verb{[91]} unspecified asian indian language
\item \verb{[92]} cebuano
\item \verb{[93]} macedonian
\item \verb{[94]} patois
\item \verb{[95]} pangasinan
\item \verb{[96]} ilocano
\item \verb{[97]} marathi
\item \verb{[98]} kashmiri
\item \verb{[99]} tamil
\item \verb{[100]} akan
\item \verb{[101]} ethiopian
\item \verb{[102]} georgian
\item \verb{[103]} yugoslavian
\item \verb{[104]} irish
\item \verb{[105]} catalan
\item \verb{[106]} mixtec
\item \verb{[107]} garifuna
\item \verb{[108]} fang
\item \verb{[109]} apache
\item \verb{[110]} bassa
\item \verb{[111]} efik
\item \verb{[112]} twi
\item \verb{[113]} maltese
\item \verb{[114]} zuni
\item \verb{[115]} ga
\item \verb{[116]} mina
\item \verb{[117]} surinamese
\item \verb{[118]} yaqui
\item \verb{[119]} quechua
\item \verb{[120]} tok pisin
\item \verb{[121]} pima
\item \verb{[122]} lingala
\item \verb{[123]} otomi
\item \verb{[124]} chindali
\item \verb{[125]} tibetan
\item \verb{[126]} oromo
\item \verb{[127]} nepali
\item \verb{[128]} eskimo
\item \verb{[129]} mam
\item \verb{[130]} choctaw
\item \verb{[131]} ashanti
\item \verb{[132]} burmese
\item \verb{[133]} hewa
\item \verb{[134]} sango
\item \verb{[135]} chitumbuka
\item \verb{[136]} maratsi
\item \verb{[137]} yemba
\item \verb{[138]} k'iche'
\item \verb{[139]} fante
\item \verb{[140]} sinhalase
\item \verb{[141]} icelandic
\item \verb{[142]} maricopa
\item \verb{[143]} quechan
\item \verb{[144]} pohnpeian
\item \verb{[145]} luganda
\item \verb{[146]} shona
\item \verb{[147]} urhobo
\item \verb{[148]} colville tribe
\item \verb{[149]} chickasau
\item \verb{[150]} dakota
\item \verb{[151]} hokkien
\item \verb{[152]} wolof
\item \verb{[153]} ewe
\item \verb{[154]} igbo
\item \verb{[155]} luhya
\item \verb{[156]} black feet
\item \verb{[157]} inupiaq
\item \verb{[158]} paiute
\item \verb{[159]} afrikaans
\item \verb{[160]} kaien
\item \verb{[161]} shawnee
\item \verb{[162]} welsh
\item \verb{[163]} chuj
\item \verb{[164]} oremagna
\item \verb{[165]} kaugel
\item \verb{[166]} sanskrit
\item \verb{[167]} hausa
\item \verb{[168]} aramaic
\item \verb{[169]} chuukese
\item \verb{[170]} dagbani
\item \verb{[171]} fijian
\item \verb{[172]} kapampagng
\item \verb{[173]} kinyarwanda
\item \verb{[174]} doric
\item \verb{[175]} sesotho
\item \verb{[176]} sousou
\item \verb{[177]} estonian
\item \verb{[178]} ebon
\item \verb{[179]} chibemba
\item \verb{[180]} chewa
\item \verb{[181]} gusii
\item \verb{[182]} gun
\item \verb{[183]} nuer
\item \verb{[184]} oro
\item \verb{[185]} kamba
\item \verb{[186]} huichol
\item \verb{[187]} papiamento
\item \verb{[188]} hiligaynon
\item \verb{[189]} mongolian
\item \verb{[190]} bunong
\item \verb{[191]} kikuyu
\item \verb{[192]} konkani
\item \verb{[193]} mende
\item \verb{[194]} naija
\item \verb{[195]} wendish
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8245/vshow}{GSS Data Explorer page for \code{othlang1_0022}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab spanish \tab polish \tab french \tab hebrew \tab russian \tab tagalog \tab chinese \tab italian \tab german \tab hungarian \tab turkish \tab korean \tab navajo \tab arabic \tab farsi/persian \tab serbian \tab lithuanian \tab slovak \tab japanese \tab portuguese \tab yiddish \tab creole \tab vietnamese \tab hindi \tab swedish \tab sign language \tab tongan \tab armenian \tab ibo \tab dutch \tab cherokee \tab samoan \tab greek \tab african \tab ilokand \tab assyrian \tab danish \tab laotian \tab finnish \tab mohawk \tab tamil \tab iap \tab english \tab urdu \tab romanian \tab ukranian \tab malayam \tab thai \tab gujarati \tab norwegian \tab latin \tab czech \tab yoruba \tab lakota \tab amharic \tab bengali \tab oriya \tab hawaiian \tab hmong \tab swahili \tab telugu \tab bulgarian \tab punjabi \tab croatian \tab romani \tab belgian \tab mandingo \tab somali \tab other native american language \tab krio \tab unspecified asian indian language \tab malaysian \tab macedonian \tab patois \tab no answer \tab khmer \tab akan \tab ethiopian \tab georgian \tab yugoslavian \tab bosnian \tab latvian \tab ilocano \tab bassa \tab efik \tab twi \tab maltese \tab pashtu \tab dari \tab kannada \tab mixtec \tab quechua \tab tok pisin \tab pima \tab lingala \tab otomi \tab chindali \tab tibetan \tab oromo \tab nepali \tab eskimo \tab mam \tab choctaw \tab chamorro \tab indonesian \tab jamaican \tab tigrinya \tab irish \tab k'iche' \tab fante \tab sinhalase \tab icelandic \tab maricopa \tab quechan \tab pohnpeian \tab luganda \tab shona \tab urhobo \tab colville tribe \tab chickasau \tab dakota \tab don't know \tab zuni \tab ashanti \tab burmese \tab igbo \tab inupiaq \tab paiute \tab kaien \tab shawnee \tab welsh \tab chuj \tab slovenian \tab afrikaans \tab chuukese \tab dagbani \tab kinyarwanda \tab doric \tab sesotho \tab sousou \tab skipped on web \tab estonian \tab chibemba \tab nuer \tab Total \cr
   2000 \tab - \tab 178 \tab 5 \tab 56 \tab 1 \tab 6 \tab 12 \tab 5 \tab 9 \tab 32 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2455 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2006 \tab - \tab 352 \tab 13 \tab 69 \tab 8 \tab 9 \tab 9 \tab 20 \tab 19 \tab 47 \tab - \tab 1 \tab 8 \tab - \tab 6 \tab 2 \tab - \tab - \tab 3 \tab 12 \tab 9 \tab 6 \tab 1 \tab 4 \tab 10 \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 6 \tab 3837 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 275 \tab 3 \tab 63 \tab 10 \tab 6 \tab 7 \tab 11 \tab 18 \tab 37 \tab - \tab 1 \tab 5 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 7 \tab 1 \tab 3 \tab 6 \tab 8 \tab - \tab 9 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1470 \tab 29 \tab 4 \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 266 \tab 4 \tab 50 \tab 6 \tab 9 \tab 6 \tab 15 \tab 16 \tab 48 \tab - \tab 1 \tab 3 \tab - \tab 5 \tab 4 \tab 1 \tab 1 \tab - \tab 12 \tab 6 \tab - \tab 5 \tab 4 \tab 6 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1508 \tab 28 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 281 \tab 7 \tab 56 \tab 2 \tab 7 \tab 7 \tab 17 \tab 10 \tab 47 \tab - \tab - \tab 7 \tab - \tab 3 \tab - \tab 4 \tab - \tab 1 \tab 5 \tab 10 \tab - \tab 5 \tab 5 \tab 9 \tab 1 \tab 6 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1397 \tab 27 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 18 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 391 \tab 3 \tab 78 \tab 2 \tab 7 \tab 10 \tab 17 \tab 22 \tab 55 \tab - \tab 2 \tab 6 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab 10 \tab 4 \tab - \tab 8 \tab 1 \tab 9 \tab 1 \tab 8 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1789 \tab 43 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 421 \tab 6 \tab 84 \tab 10 \tab 10 \tab 11 \tab 17 \tab 23 \tab 58 \tab 3 \tab 3 \tab 2 \tab 3 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 14 \tab 9 \tab 1 \tab 7 \tab 4 \tab 10 \tab - \tab 16 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2045 \tab 29 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 8 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 387 \tab 4 \tab 55 \tab 7 \tab 12 \tab 6 \tab 17 \tab 14 \tab 40 \tab 2 \tab - \tab 6 \tab 1 \tab 16 \tab 2 \tab 2 \tab - \tab 1 \tab 6 \tab 10 \tab - \tab 7 \tab 2 \tab 11 \tab 1 \tab 14 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1625 \tab 41 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab 6 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 497 \tab 5 \tab 95 \tab 7 \tab 16 \tab 5 \tab 32 \tab 24 \tab 72 \tab 3 \tab 2 \tab 19 \tab 4 \tab 12 \tab 1 \tab - \tab 1 \tab - \tab 20 \tab 16 \tab - \tab 5 \tab 4 \tab 16 \tab 1 \tab 14 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 2862 \tab 104 \tab 4 \tab 1 \tab 2 \tab 6 \tab 4 \tab 3 \tab 4 \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 30 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 74 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 494 \tab 10 \tab 51 \tab 9 \tab 12 \tab 15 \tab 26 \tab 12 \tab 46 \tab 1 \tab 2 \tab 15 \tab 3 \tab 11 \tab 4 \tab 2 \tab 1 \tab - \tab 8 \tab 11 \tab 1 \tab 8 \tab 8 \tab 14 \tab 1 \tab 11 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2564 \tab 85 \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 42 \tab 1 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 0 \tab 3542 \tab 60 \tab 657 \tab 62 \tab 94 \tab 88 \tab 177 \tab 167 \tab 482 \tab 10 \tab 13 \tab 72 \tab 13 \tab 69 \tab 20 \tab 10 \tab 7 \tab 9 \tab 94 \tab 84 \tab 10 \tab 50 \tab 43 \tab 96 \tab 8 \tab 89 \tab 2 \tab 8 \tab 3 \tab 23 \tab 4 \tab 3 \tab 38 \tab 6 \tab 2 \tab 3 \tab 5 \tab 7 \tab 7 \tab 1 \tab 26 \tab 21552 \tab 389 \tab 22 \tab 21 \tab 12 \tab 10 \tab 11 \tab 12 \tab 13 \tab 16 \tab 7 \tab 11 \tab 5 \tab 9 \tab 13 \tab 1 \tab 11 \tab 14 \tab 8 \tab 13 \tab 5 \tab 10 \tab 4 \tab 1 \tab 2 \tab 2 \tab 5 \tab 7 \tab 2 \tab 9 \tab 3 \tab 1 \tab 6 \tab 75 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 14 \tab 1 \tab 3 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 7 \tab 2 \tab 5 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 116 \tab 1 \tab 1 \tab 1 \tab 28697 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{othlang2_0022}}

Other Language: 
\code{\link{othlang2_0022}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
