% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{othlang2}
\alias{othlang2}
\title{What other languages does r speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
othlang2
}
\details{
Question C. What other language(s) do you speak?  Second response.
}
\section{Values}{

\itemize{
\item \verb{[1]} english
\item \verb{[2]} spanish
\item \verb{[3]} polish
\item \verb{[4]} french
\item \verb{[5]} hebrew
\item \verb{[6]} russian
\item \verb{[7]} tagalog
\item \verb{[8]} chinese
\item \verb{[9]} urdu
\item \verb{[10]} italian
\item \verb{[11]} romanian
\item \verb{[12]} german
\item \verb{[13]} pashtu
\item \verb{[14]} chamorro
\item \verb{[15]} ukranian
\item \verb{[16]} hungarian
\item \verb{[17]} turkish
\item \verb{[18]} khmer
\item \verb{[19]} korean
\item \verb{[20]} navajo
\item \verb{[21]} malayam
\item \verb{[22]} arabic
\item \verb{[23]} farsi/persian
\item \verb{[24]} serbian
\item \verb{[25]} lithuanian
\item \verb{[27]} thai
\item \verb{[28]} gujarati
\item \verb{[30]} dari
\item \verb{[31]} slovak
\item \verb{[32]} japanese
\item \verb{[33]} portuguese
\item \verb{[34]} vajun
\item \verb{[35]} yiddish
\item \verb{[36]} creole
\item \verb{[37]} vietnamese
\item \verb{[38]} norwegian
\item \verb{[39]} malaysian
\item \verb{[40]} latin
\item \verb{[41]} czech
\item \verb{[42]} hindi
\item \verb{[43]} hakka
\item \verb{[44]} swedish
\item \verb{[45]} vesaya
\item \verb{[46]} sign language
\item \verb{[47]} tongan
\item \verb{[48]} armenian
\item \verb{[49]} ibo
\item \verb{[50]} dutch
\item \verb{[51]} cherokee
\item \verb{[52]} samoan
\item \verb{[53]} yoruba
\item \verb{[54]} greek
\item \verb{[55]} lakota
\item \verb{[56]} african
\item \verb{[57]} anishnabee
\item \verb{[58]} ilokand
\item \verb{[59]} assyrian
\item \verb{[60]} danish
\item \verb{[61]} laotian
\item \verb{[62]} finnish
\item \verb{[63]} mohawk
\item \verb{[64]} amharic
\item \verb{[65]} bengali
\item \verb{[66]} oriya
\item \verb{[67]} hawaiian
\item \verb{[68]} albanian
\item \verb{[69]} hmong
\item \verb{[70]} indonesian
\item \verb{[71]} swahili
\item \verb{[72]} telugu
\item \verb{[73]} kannada
\item \verb{[74]} bulgarian
\item \verb{[75]} guarani
\item \verb{[76]} bosnian
\item \verb{[77]} punjabi
\item \verb{[78]} latvian
\item \verb{[79]} basque
\item \verb{[80]} bikol/bicolano
\item \verb{[81]} croatian
\item \verb{[82]} romani
\item \verb{[83]} slovenian
\item \verb{[84]} belgian
\item \verb{[85]} mandingo
\item \verb{[86]} jamaican
\item \verb{[87]} somali
\item \verb{[88]} other native american language
\item \verb{[89]} tigrinya
\item \verb{[90]} krio
\item \verb{[91]} unspecified asian indian language
\item \verb{[92]} cebuano
\item \verb{[93]} macedonian
\item \verb{[94]} patois
\item \verb{[95]} pangasinan
\item \verb{[96]} ilocano
\item \verb{[97]} marathi
\item \verb{[98]} kashmiri
\item \verb{[99]} tamil
\item \verb{[100]} akan
\item \verb{[101]} ethiopian
\item \verb{[102]} georgian
\item \verb{[103]} yugoslavian
\item \verb{[104]} irish
\item \verb{[105]} catalan
\item \verb{[106]} mixtec
\item \verb{[107]} garifuna
\item \verb{[108]} fang
\item \verb{[109]} apache
\item \verb{[110]} bassa
\item \verb{[111]} efik
\item \verb{[112]} twi
\item \verb{[113]} maltese
\item \verb{[114]} zuni
\item \verb{[115]} ga
\item \verb{[116]} mina
\item \verb{[117]} surinamese
\item \verb{[118]} yaqui
\item \verb{[119]} quechua
\item \verb{[120]} tok pisin
\item \verb{[121]} pima
\item \verb{[122]} lingala
\item \verb{[123]} otomi
\item \verb{[124]} chindali
\item \verb{[125]} tibetan
\item \verb{[126]} oromo
\item \verb{[127]} nepali
\item \verb{[128]} eskimo
\item \verb{[129]} mam
\item \verb{[130]} choctaw
\item \verb{[131]} ashanti
\item \verb{[132]} burmese
\item \verb{[133]} hewa
\item \verb{[134]} sango
\item \verb{[135]} chitumbuka
\item \verb{[136]} maratsi
\item \verb{[137]} yemba
\item \verb{[138]} k'iche'
\item \verb{[139]} fante
\item \verb{[140]} sinhalase
\item \verb{[141]} icelandic
\item \verb{[142]} maricopa
\item \verb{[143]} quechan
\item \verb{[144]} pohnpeian
\item \verb{[145]} luganda
\item \verb{[146]} shona
\item \verb{[147]} urhobo
\item \verb{[148]} colville tribe
\item \verb{[149]} chickasau
\item \verb{[150]} dakota
\item \verb{[151]} hokkien
\item \verb{[152]} wolof
\item \verb{[153]} ewe
\item \verb{[154]} igbo
\item \verb{[155]} luhya
\item \verb{[156]} black feet
\item \verb{[157]} inupiaq
\item \verb{[158]} paiute
\item \verb{[159]} afrikaans
\item \verb{[160]} kaien
\item \verb{[161]} shawnee
\item \verb{[162]} welsh
\item \verb{[163]} chuj
\item \verb{[164]} oremagna
\item \verb{[165]} kaugel
\item \verb{[166]} sanskrit
\item \verb{[167]} hausa
\item \verb{[168]} aramaic
\item \verb{[169]} chuukese
\item \verb{[170]} dagbani
\item \verb{[171]} fijian
\item \verb{[172]} kapampagng
\item \verb{[173]} kinyarwanda
\item \verb{[174]} doric
\item \verb{[175]} sesotho
\item \verb{[176]} sousou
\item \verb{[177]} estonian
\item \verb{[178]} ebon
\item \verb{[179]} chibemba
\item \verb{[180]} chewa
\item \verb{[181]} gusii
\item \verb{[182]} gun
\item \verb{[183]} nuer
\item \verb{[184]} oro
\item \verb{[185]} kamba
\item \verb{[186]} huichol
\item \verb{[187]} papiamento
\item \verb{[188]} hiligaynon
\item \verb{[189]} mongolian
\item \verb{[190]} bunong
\item \verb{[191]} kikuyu
\item \verb{[192]} konkani
\item \verb{[193]} mende
\item \verb{[194]} naija
\item \verb{[195]} wendish
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
