% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{othplan}
\alias{othplan}
\title{Other name for insurance or hmo policy}
\source{
General Social Survey https://gss.norc.org
}
\description{
othplan
}
\details{
Question Other name for insurance or HMO policy?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, another name
\item \verb{[2]} no other name
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3905/vshow}{GSS Data Explorer page for \code{othplan}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, another name \tab no other name \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1371 \tab 129 \tab 441 \tab 76 \tab 6 \tab 2023 \cr
   Total \tab 1371 \tab 129 \tab 441 \tab 76 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{natdeg}},
\code{\link{typevisa}},
\code{\link{visa}}
}
\concept{Global Economics}
\keyword{variable}
