% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{othreasn}
\alias{othreasn}
\title{How important to try to undr'stnd reasoning of othr o}
\source{
General Social Survey https://gss.norc.org
}
\description{
othreasn
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
F. To try to understand the reasoning of people with other opinions
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4912/vshow}{GSS Data Explorer page for \code{othreasn}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 43 \tab 19 \tab 39 \tab 124 \tab 263 \tab 324 \tab 654 \tab 4 \tab 2 \tab 2812 \cr
   2014 \tab 1274 \tab 45 \tab 22 \tab 23 \tab 109 \tab 205 \tab 280 \tab 564 \tab 15 \tab 1 \tab 2538 \cr
   Total \tab 2614 \tab 88 \tab 41 \tab 62 \tab 233 \tab 468 \tab 604 \tab 1218 \tab 19 \tab 3 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{paytaxes}},
\code{\link{relmeet}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
