% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_580.R
\name{othrel}
\alias{othrel}
\title{Attend some other type of religious event}
\source{
General Social Survey https://gss.norc.org
}
\description{
othrel
}
\details{
Question During the last seven days did you do the following:
C. Attend some other type of religious event or meeting (e.g. prayer breakfasts, Bible study groups, choir practices, church sponsored lectures, adult fellowship meetings)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5320/vshow}{GSS Data Explorer page for \code{othrel}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1996 \tab 1953 \tab 97 \tab 784 \tab 70 \tab - \tab 2904 \cr
   1998 \tab 1387 \tab 180 \tab 1223 \tab 41 \tab 1 \tab 2832 \cr
   Total \tab 3340 \tab 277 \tab 2007 \tab 111 \tab 1 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attreg}},
\code{\link{friday}},
\code{\link{mediarel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
