% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{othrel3}
\alias{othrel3}
\title{What type of religious events or meetings}
\source{
General Social Survey https://gss.norc.org
}
\description{
othrel3
}
\details{
Question (Besides the regular, weekly worship service,) What type of religious events or meetings did you attend during the last seven days?
3. Third mention
}
\section{Values}{

\itemize{
\item \code{1} bible study,class
\item \code{2} choir, choral, concert, play
\item \code{3} discussion group
\item \code{4} lecture
\item \code{5} governing board, etc
\item \code{6} preaching, taught sunday school
\item \code{7} prayer group, breakfast, meeting
\item \code{8} counseling
\item \code{9} worship service
\item \code{10} other, religious
\item \code{11} other, non-religious
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5323/vshow}{GSS Data Explorer page for \code{othrel3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab choir, choral, concert, play \tab governing board, etc \tab preaching, taught sunday school \tab prayer group, breakfast, meeting \tab worship service \tab other, religious \tab Total \cr
   1996 \tab 2898 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 2829 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2832 \cr
   Total \tab 5727 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
