% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{othresp}
\alias{othresp}
\title{How resp. person who you were angry at?}
\source{
General Social Survey https://gss.norc.org
}
\description{
othresp
}
\details{
Question Think again about the situation that made you angry or annoyed. How responsible was the person/people who you were angry or annoyed at for the situation?  Think of a scale where 0 means he/she was/were not at all responsible, while 10 means he/she/they was/were completely responsible.
}
\section{Values}{

\itemize{
\item \code{0} not at all responsible
\item \code{10} completely responsible
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1784/vshow}{GSS Data Explorer page for \code{othresp}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab not at all responsible \tab 1 \tab completely responsible \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1996 \tab 1908 \tab 43 \tab 12 \tab 458 \tab 16 \tab 17 \tab 22 \tab 78 \tab 43 \tab 81 \tab 114 \tab 82 \tab 30 \tab 2904 \cr
   Total \tab 1908 \tab 43 \tab 12 \tab 458 \tab 16 \tab 17 \tab 22 \tab 78 \tab 43 \tab 81 \tab 114 \tab 82 \tab 30 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{notam}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
