% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{othsch}
\alias{othsch}
\title{Children attend other school}
\source{
General Social Survey https://gss.norc.org
}
\description{
othsch
}
\details{
Question Do/did all of your children attend public school for all of their elementary and high school education?
A. Which types of school do/did your children attend?
6. Other
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1154/vshow}{GSS Data Explorer page for \code{othsch}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 2634 \tab 16 \tab 139 \tab 43 \tab - \tab 2832 \cr
   2000 \tab 2540 \tab 20 \tab 229 \tab 28 \tab - \tab 2817 \cr
   Total \tab 5174 \tab 36 \tab 368 \tab 71 \tab 0 \tab 5649 \cr
}
}

\seealso{
Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{madenkid}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numbwfrd}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcand}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}

Other Children: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adoption}},
\code{\link{adoptkid}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{aidkids}},
\code{\link{allergic}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{cathsch}},
\code{\link{chembal}},
\code{\link{childs}},
\code{\link{chldben}},
\code{\link{chldcare}},
\code{\link{chldeduc}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldhrt}},
\code{\link{chldidel}},
\code{\link{chldidel_next}},
\code{\link{chldlove}},
\code{\link{chldmore}},
\code{\link{chldmorl}},
\code{\link{chldnum}},
\code{\link{chldpref_next}},
\code{\link{chldprob}},
\code{\link{chldsafe}},
\code{\link{chldsch}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldskls}},
\code{\link{chldsoon}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{chldvig}},
\code{\link{chrissch}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{comknows}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{discipln}},
\code{\link{eldersup}},
\code{\link{evkid}},
\code{\link{failure}},
\code{\link{famper4}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{fourkids}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{helpoth}},
\code{\link{hlth8}},
\code{\link{homesch}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{illphys}},
\code{\link{impkids}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpicked}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kid5up}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidhlth}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidlived}},
\code{\link{kidneedy}},
\code{\link{kidnofre}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidnum}},
\code{\link{kidout}},
\code{\link{kidpars}},
\code{\link{kidresp}},
\code{\link{kidskips}},
\code{\link{kidsocst}},
\code{\link{kidssol}},
\code{\link{kidtrble}},
\code{\link{kidxfilm}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{law7}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{mapaid}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{mntlill}},
\code{\link{morextrm}},
\code{\link{mykids}},
\code{\link{nextdoor}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obeythnk}},
\code{\link{obrespct}},
\code{\link{onekid}},
\code{\link{oth3}},
\code{\link{othkids}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{popular}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{raised}},
\code{\link{relsch}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{sellbaby}},
\code{\link{spanking}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spendeve}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}},
\code{\link{spwrkpar}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{stressfl}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{sufadult}},
\code{\link{thnkself}},
\code{\link{threkids}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{trbllaw}},
\code{\link{twokids}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{workhard}},
\code{\link{zombies}}

Other Education: 
\code{\link{actgreek}},
\code{\link{actschl}},
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{bilinged}},
\code{\link{cathsch}},
\code{\link{chldeduc}},
\code{\link{chrissch}},
\code{\link{classics}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{coldeg1}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{colop}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{datesch}},
\code{\link{degree}},
\code{\link{denomsch}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{educ}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{educemp}},
\code{\link{educop}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{govedop}},
\code{\link{homesch}},
\code{\link{hsbasics}},
\code{\link{hsbio}},
\code{\link{hscaring}},
\code{\link{hschem}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsmath}},
\code{\link{hsorder}},
\code{\link{hsphys}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inteduc}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{majorcol}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memschl}},
\code{\link{natdeg}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{opeduc}},
\code{\link{oppared}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{relsch}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{schleth}},
\code{\link{schlhead}},
\code{\link{schoolng}},
\code{\link{scimath}},
\code{\link{sexeduc}},
\code{\link{solgreek}},
\code{\link{solschl}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{speduc}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{teacheth}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{trstprof}},
\code{\link{undoccol}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{whoteach}}
}
\concept{Children}
\concept{Education}
\concept{Religion}
\keyword{variable}
