% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{othself}
\alias{othself}
\title{Was this prob of narrow or wide concern}
\source{
General Social Survey https://gss.norc.org
}
\description{
othself
}
\details{
Question What about some representatives or governmental officials outside of the local community -- on the (county, IF LOCAL UNIT BELOW COUNTY LEVEL), state, or national level? Have you ever contacted or written to such a person on some need or problem?
A. Was this need or problem primarily of concern to you, your friends, and family or was it an issue of wider concern?
}
\section{Values}{

\itemize{
\item \code{1} self frnds famly
\item \code{2} both
\item \code{3} wider concern
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1034/vshow}{GSS Data Explorer page for \code{othself}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
