% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{othshelp}
\alias{othshelp}
\title{People should help less fortunate others}
\source{
General Social Survey https://gss.norc.org
}
\description{
othshelp
}
\details{
Question Please tell me whether you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree with the following statements:
A. People should be willing to help others who are less fortunate
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2890/vshow}{GSS Data Explorer page for \code{othshelp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 567 \tab 638 \tab 123 \tab 15 \tab 9 \tab 12 \tab 8 \tab 2765 \cr
   2004 \tab 1472 \tab 583 \tab 625 \tab 105 \tab 10 \tab 13 \tab 2 \tab 2 \tab 2812 \cr
   2012 \tab 666 \tab 510 \tab 678 \tab 89 \tab 15 \tab 4 \tab 6 \tab 6 \tab 1974 \cr
   2014 \tab 1264 \tab 609 \tab 549 \tab 77 \tab 19 \tab 10 \tab 3 \tab 7 \tab 2538 \cr
   Total \tab 4795 \tab 2269 \tab 2490 \tab 394 \tab 59 \tab 36 \tab 23 \tab 23 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{peoptrbl}},
\code{\link{selffrst}},
\code{\link{selfless}}
}
\concept{Altruism}
\keyword{variable}
