% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{otjtrain}
\alias{otjtrain}
\title{Received any on-the-job training}
\source{
General Social Survey https://gss.norc.org
}
\description{
otjtrain
}
\details{
Question A. Did you receive any on-the-job training (anywhere) about this kind of (product/service) or its production process?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no/prcticly none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2149/vshow}{GSS Data Explorer page for \code{otjtrain}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no/prcticly none \tab no answer \tab Total \cr
   1998 \tab 1455 \tab 920 \tab 428 \tab 29 \tab 2832 \cr
   Total \tab 1455 \tab 920 \tab 428 \tab 29 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}
}
\concept{Jobs}
\keyword{variable}
