% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{outofbiz}
\alias{outofbiz}
\title{The organization went out of business}
\source{
General Social Survey https://gss.norc.org
}
\description{
outofbiz
}
\details{
Question In the last five years, did any of the following things happen to an organization that you worked for?
A. The organization went out of business or closed down?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1267/vshow}{GSS Data Explorer page for \code{outofbiz}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab - \tab 149 \tab 1303 \tab 65 \tab - \tab 1517 \cr
   Total \tab 0 \tab 149 \tab 1303 \tab 65 \tab 0 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{afailure}},
\code{\link{automatn}},
\code{\link{baseofop}},
\code{\link{chngeorg}},
\code{\link{cowrkrel}},
\code{\link{drugtest}},
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{futpromo}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{helporg}},
\code{\link{idecide}},
\code{\link{imatter}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobcntrl}},
\code{\link{jobjudge}},
\code{\link{jobvshme}},
\code{\link{jobyears}},
\code{\link{learning}},
\code{\link{levels}()},
\code{\link{lottosay}},
\code{\link{ltrsref}},
\code{\link{merged}},
\code{\link{monitred}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{offsup}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{orgfin}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{othpay}},
\code{\link{pessimst}},
\code{\link{physical}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{proudorg}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{racpromo}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{reorg}},
\code{\link{samejob}},
\code{\link{samevals}},
\code{\link{satself}},
\code{\link{schoolng}},
\code{\link{sexpromo}},
\code{\link{skiltest}},
\code{\link{slfrspct}},
\code{\link{splocnum}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{totunemp}},
\code{\link{training}},
\code{\link{unmanrel}},
\code{\link{unpower}},
\code{\link{unprog}},
\code{\link{unvote}},
\code{\link{wherewrk}},
\code{\link{wrkindep}},
\code{\link{wrkmuch}},
\code{\link{wrkwell}},
\code{\link{wrkyears}},
\code{\link{youmoney}},
\code{\link{yousup}}

Other Corporations: 
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{equal2}},
\code{\link{equal6}},
\code{\link{merged}},
\code{\link{orgfin}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{startbiz}},
\code{\link{whybiz}},
\code{\link{yearbiz}}

Other Trauma: 
\code{\link{clsclsex}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}},
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}},
\code{\link{oth5}},
\code{\link{ownclerg}},
\code{\link{prob1}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unemp5}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{work1}},
\code{\link{work10}},
\code{\link{work2}},
\code{\link{work3}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}
}
\concept{Corporations}
\concept{Trauma}
\concept{Work Organizations}
\keyword{variable}
