% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_182.R
\name{outraged}
\alias{outraged}
\title{Felt outraged at something somebody}
\source{
General Social Survey https://gss.norc.org
}
\description{
outraged
}
\details{
Question Now I'm going to read a list of different feelings that people sometimes have. After each one, I would like you to tell me on how many days you have felt this way during the past 7 days. On how many days in the past 7 days have you... (SPECIFY NUMBER OF DAYS)
(IF RESPONDENT ASKS, SAY THAT FEELING THIS WAY AT ANY TIME DURING A DAY MAKES THAT DAY "COUNT.") C. Felt outraged at something somebody had done?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1708/vshow}{GSS Data Explorer page for \code{outraged}}.
}

\seealso{
Other Emotions: 
\code{\link{ashamed}},
\code{\link{calm}},
\code{\link{hapfeel}},
\code{\link{sad}},
\code{\link{shakeblu}}
}
\concept{Emotions}
\keyword{variable}
