% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{overwork}
\alias{overwork}
\title{R has too much work to do well}
\source{
General Social Survey https://gss.norc.org
}
\description{
overwork
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
F. I have too much work to do everything well
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2779/vshow}{GSS Data Explorer page for \code{overwork}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 113 \tab 367 \tab 1083 \tab 206 \tab 14 \tab 13 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 98 \tab 361 \tab 1063 \tab 189 \tab 10 \tab 13 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 93 \tab 285 \tab 668 \tab 114 \tab 4 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 100 \tab 321 \tab 732 \tab 89 \tab 1 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 127 \tab 369 \tab 786 \tab 131 \tab 5 \tab 30 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 185 \tab 472 \tab 1100 \tab 195 \tab 14 \tab 12 \tab 20 \tab 3544 \cr
   Total \tab 8336 \tab 716 \tab 2175 \tab 5432 \tab 924 \tab 48 \tab 98 \tab 20 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
