% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_331.R
\name{ovrmedkd}
\alias{ovrmedkd}
\title{Docs are over medicating child with common problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
ovrmedkd
}
\details{
Question Now, I''d like to ask you for your opinions on several issues. For each question, please tell me if you strongly agree, agree somewhat, disagree somewhat or strongly disagree with the statement.
A. Doctors today are over-medicating children with common behavior problems.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2961/vshow}{GSS Data Explorer page for \code{ovrmedkd}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree somewhat \tab disagree somewhat \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 579 \tab 529 \tab 146 \tab 37 \tab 90 \tab 12 \tab 2765 \cr
   Total \tab 1372 \tab 579 \tab 529 \tab 146 \tab 37 \tab 90 \tab 12 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{comknows}},
\code{\link{failure}},
\code{\link{illphys}},
\code{\link{medkdneg}},
\code{\link{outsider}},
\code{\link{putsoff}},
\code{\link{sufadult}},
\code{\link{violpeop}},
\code{\link{violself}}
}
\concept{Mental Health II}
\keyword{variable}
