% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{ownclerg}
\alias{ownclerg}
\title{Did this happen with a leader in congregation r was attending}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownclerg
}
\details{
Question Did this happen with a leader in a congregation you were yourself attending?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3937/vshow}{GSS Data Explorer page for \code{ownclerg}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 1980 \tab 24 \tab 19 \tab 2023 \cr
   Total \tab 1980 \tab 24 \tab 19 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{harsexcl}},
\code{\link{harsexjb}}
}
\concept{Clergy/Sex}
\keyword{variable}
