% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{owngun}
\alias{owngun}
\title{Have gun in home}
\source{
General Social Survey https://gss.norc.org
}
\description{
owngun
}
\details{
Question Do you happen to have in your home (IF HOUSE:  or garage) any guns or revolvers?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
   A/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024 \cr
   Full \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/679/vshow}{GSS Data Explorer page for \code{owngun}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab refused \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1973 \tab - \tab 707 \tab 773 \tab 15 \tab 9 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 684 \tab 785 \tab 10 \tab 4 \tab 1 \tab - \tab 1484 \cr
   1976 \tab - \tab 697 \tab 779 \tab 17 \tab 6 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 771 \tab 748 \tab 2 \tab 7 \tab 2 \tab - \tab 1530 \cr
   1980 \tab - \tab 695 \tab 760 \tab 2 \tab 10 \tab 1 \tab - \tab 1468 \cr
   1982 \tab - \tab 810 \tab 1019 \tab 22 \tab 6 \tab 3 \tab - \tab 1860 \cr
   1984 \tab - \tab 662 \tab 795 \tab 9 \tab 6 \tab 1 \tab - \tab 1473 \cr
   1985 \tab - \tab 678 \tab 842 \tab 10 \tab 4 \tab - \tab - \tab 1534 \cr
   1987 \tab - \tab 770 \tab 1032 \tab 10 \tab 7 \tab - \tab - \tab 1819 \cr
   1988 \tab 504 \tab 389 \tab 571 \tab 10 \tab 7 \tab - \tab - \tab 1481 \cr
   1989 \tab 504 \tab 475 \tab 555 \tab - \tab 3 \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 387 \tab 520 \tab - \tab 9 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 393 \tab 583 \tab 10 \tab 7 \tab - \tab - \tab 1517 \cr
   1993 \tab 531 \tab 452 \tab 614 \tab 7 \tab 2 \tab - \tab - \tab 1606 \cr
   1994 \tab 996 \tab 810 \tab 1160 \tab 19 \tab 4 \tab 3 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 772 \tab 1142 \tab 7 \tab - \tab 2 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 654 \tab 1215 \tab 6 \tab 6 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 603 \tab 1231 \tab 23 \tab 3 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 310 \tab 605 \tab 9 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 319 \tab 558 \tab 16 \tab 3 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 664 \tab 1307 \tab 30 \tab - \tab 2 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 464 \tab 870 \tab 15 \tab - \tab 3 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 394 \tab 839 \tab 42 \tab - \tab 6 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 440 \tab 841 \tab 24 \tab - \tab 3 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 529 \tab 1128 \tab 54 \tab - \tab 3 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 593 \tab 1246 \tab 50 \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 537 \tab 993 \tab 39 \tab - \tab 5 \tab - \tab 2348 \cr
   2021 \tab - \tab 1383 \tab 2529 \tab 10 \tab - \tab 3 \tab 107 \tab 4032 \cr
   2022 \tab 1203 \tab 758 \tab 1511 \tab 44 \tab - \tab 3 \tab 25 \tab 3544 \cr
   Total \tab 18541 \tab 17800 \tab 27551 \tab 512 \tab 103 \tab 47 \tab 132 \tab 64686 \cr
}
}

\seealso{
Other Core: 
\code{\link{burglr}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{pistol}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{rifle}},
\code{\link{robbry}},
\code{\link{shotgun}}
}
\concept{Core}
\keyword{variable}
