% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{ownhh1}
\alias{ownhh1}
\title{Aged should stop living with parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownhh1
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important is it For them to be...
No longer living in their parents' household
A. By what age should this normally occur?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2899/vshow}{GSS Data Explorer page for \code{ownhh1}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab 13 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 30 \tab 45 \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1635 \tab 1 \tab 2 \tab 8 \tab 253 \tab 72 \tab 138 \tab 215 \tab 143 \tab 68 \tab 52 \tab 142 \tab 3 \tab 3 \tab 8 \tab 4 \tab 1 \tab 16 \tab 1 \tab 2765 \cr
   Total \tab 1635 \tab 1 \tab 2 \tab 8 \tab 253 \tab 72 \tab 138 \tab 215 \tab 143 \tab 68 \tab 52 \tab 142 \tab 3 \tab 3 \tab 8 \tab 4 \tab 1 \tab 16 \tab 1 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{ftwork}},
\code{\link{ftwork1}},
\code{\link{ownhh}}
}
\concept{Adult Transitions}
\keyword{variable}
