% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{ownmass}
\alias{ownmass}
\title{Govts role in local mass transportation}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownmass
}
\details{
Question What do you think the government''s role in each of these industries should be?
B. Local mass transportation.
}
\section{Values}{

\itemize{
\item \code{1} own it
\item \code{2} control prices
\item \code{3} leave alone
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4072/vshow}{GSS Data Explorer page for \code{ownmass}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab own it \tab control prices \tab leave alone \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 60 \tab 291 \tab 261 \tab 49 \tab 16 \tab 1534 \cr
   Total \tab 857 \tab 60 \tab 291 \tab 261 \tab 49 \tab 16 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
