% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{ownthing}
\alias{ownthing}
\title{Youthful protest is good}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownthing
}
\details{
Question Now I am going to read you several statements. For each one please tell me whether you agree strongly, agree, disagree, or disagree strongly. The first statement is:
A. It is wonderful that young people today have greater freedom to protest against things they don''t like and to ''do their own thing.''
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1166/vshow}{GSS Data Explorer page for \code{ownthing}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 157 \tab 753 \tab 341 \tab 68 \tab 51 \tab 2 \tab 1372 \cr
   1994 \tab 1518 \tab 201 \tab 725 \tab 404 \tab 101 \tab 31 \tab 12 \tab 2992 \cr
   Total \tab 1518 \tab 358 \tab 1478 \tab 745 \tab 169 \tab 82 \tab 14 \tab 4364 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{openmind}},
\code{\link{talkback}},
\code{\link{twoclass}}
}
\concept{Intergroup Relations}
\keyword{variable}
