% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{paborn}
\alias{paborn}
\title{Was r's father (or oldest same-sex parent) born in this country?}
\source{
General Social Survey https://gss.norc.org
}
\description{
paborn
}
\details{
Question Was your father born in this country?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7477/vshow}{GSS Data Explorer page for \code{paborn}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 49 \tab 3269 \tab 649 \tab 47 \tab 1 \tab 17 \tab 4032 \cr
   2022 \tab - \tab 2787 \tab 696 \tab 45 \tab 8 \tab 8 \tab 3544 \cr
   Total \tab 49 \tab 6056 \tab 1345 \tab 92 \tab 9 \tab 25 \tab 7576 \cr
}
}

\keyword{variable}
