% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{padeath}
\alias{padeath}
\title{Death of father}
\source{
General Social Survey https://gss.norc.org
}
\description{
padeath
}
\details{
Question A. Death of respondent's father
}
\section{Values}{

\itemize{
\item \code{0} no death reported
\item \code{1} died before r 16
\item \code{2} died 6+ yrs ago
\item \code{3} died 1-5 yrs ago
\item \code{4} died last year
\item \code{8} cant tell
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/748/vshow}{GSS Data Explorer page for \code{padeath}}.
}

\seealso{
Other Core: 
\code{\link{chlddth}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{madeath}},
\code{\link{unrel1}},
\code{\link{unrel4}}

Other Family: 
\code{\link{chlddth}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{madeath}}
}
\concept{Core}
\concept{Family}
\keyword{variable}
