% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{padotsvp}
\alias{padotsvp}
\title{Fa. D.o.t. code - specific vocat. prep}
\source{
General Social Survey https://gss.norc.org
}
\description{
padotsvp
}
\details{
Question E. Father's Occupation Specific Vocation Preparation (SVP).
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/139/vshow}{GSS Data Explorer page for \code{padotsvp}}.
}

\seealso{
Other Core: 
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotthn}},
\code{\link{partyid}},
\code{\link{pres68}},
\code{\link{spdotpre}},
\code{\link{vote68}}

Other D.O.T. Codes: 
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotthn}},
\code{\link{spdotpre}}
}
\concept{Core}
\concept{D.O.T. Codes}
\keyword{variable}
