% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{padotthn}
\alias{padotthn}
\title{Fa. D.o.t. code - relation to things}
\source{
General Social Survey https://gss.norc.org
}
\description{
padotthn
}
\details{
Question C. Father's Occupational Relationship with Things.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{partyid}},
\code{\link{pres68}},
\code{\link{spdotpre}},
\code{\link{vote68}}
}
\concept{Core}
\keyword{variable}
