% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{paeduc}
\alias{paeduc}
\title{Highest year school completed, father (or oldest same-sex parent)}
\source{
General Social Survey https://gss.norc.org
}
\description{
paeduc
}
\details{
Question RESPONDENT'S FATHER'S (FATHER SUBSTITUTE'S) EDUCATION
}
\section{Values}{

\itemize{
\item \verb{[0]} no formal schooling
\item \verb{[1]} 1st grade
\item \verb{[2]} 2nd grade
\item \verb{[3]} 3rd grade
\item \verb{[4]} 4th grade
\item \verb{[5]} 5th grade
\item \verb{[6]} 6th grade
\item \verb{[7]} 7th grade
\item \verb{[8]} 8th grade
\item \verb{[9]} 9th grade
\item \verb{[10]} 10th grade
\item \verb{[11]} 11th grade
\item \verb{[12]} 12th grade
\item \verb{[13]} 1 year of college
\item \verb{[14]} 2 years of college
\item \verb{[15]} 3 years of college
\item \verb{[16]} 4 years of college
\item \verb{[17]} 5 years of college
\item \verb{[18]} 6 years of college
\item \verb{[19]} 7 years of college
\item \verb{[20]} 8 or more years of college
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/56/vshow}{GSS Data Explorer page for \code{paeduc}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab no formal schooling \tab 1st grade \tab 2nd grade \tab 3rd grade \tab 4th grade \tab 5th grade \tab 6th grade \tab 7th grade \tab 8th grade \tab 9th grade \tab 10th grade \tab 11th grade \tab 12th grade \tab 1 year of college \tab 2 years of college \tab 3 years of college \tab 4 years of college \tab 5 years of college \tab 6 years of college \tab 7 years of college \tab 8 or more years of college \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 54 \tab 6 \tab 12 \tab 51 \tab 57 \tab 43 \tab 68 \tab 49 \tab 231 \tab 34 \tab 58 \tab 44 \tab 195 \tab 24 \tab 40 \tab 9 \tab 59 \tab 5 \tab 17 \tab 3 \tab 12 \tab 297 \tab 235 \tab 10 \tab 1613 \cr
   1973 \tab 68 \tab 8 \tab 14 \tab 52 \tab 46 \tab 39 \tab 89 \tab 45 \tab 261 \tab 27 \tab 41 \tab 44 \tab 223 \tab 14 \tab 48 \tab 11 \tab 65 \tab 6 \tab 7 \tab 5 \tab 11 \tab 216 \tab 151 \tab 13 \tab 1504 \cr
   1974 \tab 52 \tab 8 \tab 10 \tab 49 \tab 46 \tab 43 \tab 81 \tab 47 \tab 213 \tab 43 \tab 55 \tab 37 \tab 234 \tab 21 \tab 48 \tab 11 \tab 50 \tab 4 \tab 15 \tab 1 \tab 10 \tab 223 \tab 182 \tab 1 \tab 1484 \cr
   1975 \tab 55 \tab 3 \tab 13 \tab 49 \tab 50 \tab 40 \tab 92 \tab 55 \tab 231 \tab 37 \tab 55 \tab 39 \tab 240 \tab 31 \tab 44 \tab 11 \tab 53 \tab 12 \tab 11 \tab 1 \tab 11 \tab 204 \tab 151 \tab 2 \tab 1490 \cr
   1976 \tab 48 \tab 7 \tab 12 \tab 44 \tab 47 \tab 39 \tab 91 \tab 50 \tab 223 \tab 37 \tab 45 \tab 41 \tab 243 \tab 21 \tab 40 \tab 8 \tab 48 \tab 9 \tab 15 \tab 4 \tab 15 \tab 217 \tab 186 \tab 9 \tab 1499 \cr
   1977 \tab 39 \tab 2 \tab 15 \tab 45 \tab 64 \tab 37 \tab 80 \tab 49 \tab 228 \tab 35 \tab 62 \tab 38 \tab 219 \tab 20 \tab 38 \tab 13 \tab 46 \tab 5 \tab 13 \tab 7 \tab 13 \tab 267 \tab 193 \tab 2 \tab 1530 \cr
   1978 \tab 41 \tab - \tab 11 \tab 36 \tab 39 \tab 35 \tab 88 \tab 53 \tab 213 \tab 44 \tab 58 \tab 29 \tab 256 \tab 29 \tab 51 \tab 8 \tab 54 \tab 16 \tab 12 \tab 6 \tab 10 \tab 235 \tab 205 \tab 3 \tab 1532 \cr
   1980 \tab 42 \tab 3 \tab 10 \tab 29 \tab 51 \tab 36 \tab 86 \tab 36 \tab 201 \tab 38 \tab 36 \tab 36 \tab 248 \tab 29 \tab 42 \tab 13 \tab 65 \tab 13 \tab 18 \tab 9 \tab 17 \tab 178 \tab 230 \tab 2 \tab 1468 \cr
   1982 \tab 65 \tab 3 \tab 7 \tab 42 \tab 46 \tab 40 \tab 74 \tab 49 \tab 226 \tab 46 \tab 66 \tab 36 \tab 309 \tab 29 \tab 55 \tab 14 \tab 72 \tab 10 \tab 20 \tab 6 \tab 17 \tab 323 \tab 296 \tab 9 \tab 1860 \cr
   1983 \tab 41 \tab 3 \tab 12 \tab 28 \tab 32 \tab 47 \tab 68 \tab 39 \tab 202 \tab 52 \tab 53 \tab 39 \tab 302 \tab 30 \tab 57 \tab 5 \tab 102 \tab 9 \tab 22 \tab 7 \tab 18 \tab 224 \tab 207 \tab - \tab 1599 \cr
   1984 \tab 27 \tab 1 \tab 4 \tab 32 \tab 41 \tab 28 \tab 66 \tab 29 \tab 207 \tab 41 \tab 57 \tab 43 \tab 278 \tab 28 \tab 51 \tab 9 \tab 79 \tab 9 \tab 15 \tab 6 \tab 16 \tab 201 \tab 202 \tab 3 \tab 1473 \cr
   1985 \tab 31 \tab 3 \tab 14 \tab 29 \tab 36 \tab 30 \tab 77 \tab 39 \tab 194 \tab 40 \tab 67 \tab 45 \tab 288 \tab 29 \tab 64 \tab 13 \tab 88 \tab 7 \tab 26 \tab 8 \tab 25 \tab 187 \tab 194 \tab - \tab 1534 \cr
   1986 \tab 32 \tab 4 \tab 5 \tab 34 \tab 35 \tab 25 \tab 74 \tab 30 \tab 178 \tab 30 \tab 52 \tab 44 \tab 297 \tab 22 \tab 54 \tab 13 \tab 85 \tab 6 \tab 13 \tab 6 \tab 12 \tab 193 \tab 222 \tab 4 \tab 1470 \cr
   1987 \tab 33 \tab 5 \tab 8 \tab 48 \tab 44 \tab 45 \tab 83 \tab 39 \tab 223 \tab 52 \tab 56 \tab 45 \tab 318 \tab 29 \tab 63 \tab 15 \tab 98 \tab 15 \tab 19 \tab 16 \tab 28 \tab 234 \tab 291 \tab 12 \tab 1819 \cr
   1988 \tab 28 \tab 1 \tab 11 \tab 41 \tab 24 \tab 33 \tab 59 \tab 26 \tab 183 \tab 39 \tab 49 \tab 30 \tab 291 \tab 33 \tab 43 \tab 11 \tab 94 \tab 10 \tab 28 \tab 4 \tab 27 \tab 208 \tab 189 \tab 19 \tab 1481 \cr
   1989 \tab 62 \tab 4 \tab 14 \tab 43 \tab 28 \tab 35 \tab 59 \tab 36 \tab 180 \tab 40 \tab 57 \tab 33 \tab 302 \tab 33 \tab 58 \tab 16 \tab 96 \tab 10 \tab 30 \tab 7 \tab 24 \tab 172 \tab 197 \tab 1 \tab 1537 \cr
   1990 \tab 22 \tab 2 \tab 11 \tab 33 \tab 29 \tab 17 \tab 42 \tab 25 \tab 159 \tab 33 \tab 42 \tab 20 \tab 300 \tab 35 \tab 54 \tab 17 \tab 85 \tab 9 \tab 21 \tab 6 \tab 20 \tab 165 \tab 190 \tab 35 \tab 1372 \cr
   1991 \tab 17 \tab - \tab 7 \tab 31 \tab 22 \tab 22 \tab 61 \tab 27 \tab 165 \tab 39 \tab 49 \tab 38 \tab 300 \tab 28 \tab 77 \tab 12 \tab 103 \tab 12 \tab 24 \tab 13 \tab 22 \tab 211 \tab 205 \tab 32 \tab 1517 \cr
   1993 \tab 22 \tab 2 \tab 9 \tab 39 \tab 22 \tab 23 \tab 63 \tab 34 \tab 187 \tab 34 \tab 62 \tab 34 \tab 317 \tab 37 \tab 79 \tab 20 \tab 123 \tab 12 \tab 36 \tab 9 \tab 31 \tab 185 \tab 197 \tab 29 \tab 1606 \cr
   1994 \tab 43 \tab 5 \tab 17 \tab 47 \tab 35 \tab 41 \tab 109 \tab 60 \tab 307 \tab 72 \tab 112 \tab 53 \tab 637 \tab 57 \tab 161 \tab 22 \tab 216 \tab 19 \tab 62 \tab 21 \tab 59 \tab 330 \tab 492 \tab 15 \tab 2992 \cr
   1996 \tab 31 \tab 2 \tab 6 \tab 40 \tab 40 \tab 43 \tab 98 \tab 43 \tab 254 \tab 63 \tab 101 \tab 59 \tab 631 \tab 72 \tab 148 \tab 40 \tab 229 \tab 22 \tab 60 \tab 15 \tab 69 \tab 324 \tab 502 \tab 12 \tab 2904 \cr
   1998 \tab 28 \tab 4 \tab 15 \tab 45 \tab 40 \tab 42 \tab 96 \tab 53 \tab 274 \tab 58 \tab 76 \tab 56 \tab 607 \tab 58 \tab 145 \tab 26 \tab 256 \tab 31 \tab 57 \tab 16 \tab 58 \tab 275 \tab 503 \tab 13 \tab 2832 \cr
   2000 \tab 32 \tab 4 \tab 19 \tab 44 \tab 36 \tab 43 \tab 95 \tab 35 \tab 227 \tab 52 \tab 95 \tab 59 \tab 627 \tab 65 \tab 129 \tab 26 \tab 212 \tab 13 \tab 63 \tab 15 \tab 55 \tab 326 \tab 518 \tab 27 \tab 2817 \cr
   2002 \tab 35 \tab 6 \tab 19 \tab 38 \tab 28 \tab 34 \tab 87 \tab 32 \tab 234 \tab 75 \tab 123 \tab 59 \tab 661 \tab 62 \tab 158 \tab 27 \tab 210 \tab 18 \tab 61 \tab 18 \tab 50 \tab 185 \tab 526 \tab 19 \tab 2765 \cr
   2004 \tab 33 \tab 3 \tab 5 \tab 36 \tab 27 \tab 29 \tab 84 \tab 48 \tab 204 \tab 79 \tab 81 \tab 65 \tab 692 \tab 72 \tab 165 \tab 25 \tab 256 \tab 29 \tab 69 \tab 28 \tab 60 \tab 182 \tab 514 \tab 26 \tab 2812 \cr
   2006 \tab 71 \tab 11 \tab 24 \tab 65 \tab 36 \tab 40 \tab 103 \tab 48 \tab 220 \tab 70 \tab 87 \tab 56 \tab 684 \tab 83 \tab 168 \tab 35 \tab 253 \tab 28 \tab 75 \tab 25 \tab 76 \tab 185 \tab 2049 \tab 18 \tab 4510 \cr
   2008 \tab 37 \tab 1 \tab 9 \tab 37 \tab 22 \tab 26 \tab 70 \tab 22 \tab 138 \tab 41 \tab 57 \tab 48 \tab 502 \tab 56 \tab 99 \tab 22 \tab 187 \tab 15 \tab 44 \tab 16 \tab 38 \tab 177 \tab 354 \tab 5 \tab 2023 \cr
   2010 \tab 33 \tab 5 \tab 11 \tab 40 \tab 19 \tab 29 \tab 58 \tab 30 \tab 133 \tab 47 \tab 55 \tab 48 \tab 536 \tab 54 \tab 95 \tab 19 \tab 177 \tab 13 \tab 59 \tab 15 \tab 40 \tab 150 \tab 370 \tab 8 \tab 2044 \cr
   2012 \tab 36 \tab 6 \tab 11 \tab 30 \tab 22 \tab 28 \tab 51 \tab 20 \tab 122 \tab 54 \tab 60 \tab 44 \tab 475 \tab 51 \tab 95 \tab 15 \tab 190 \tab 17 \tab 44 \tab 10 \tab 59 \tab 193 \tab 338 \tab 3 \tab 1974 \cr
   2014 \tab 41 \tab 5 \tab 23 \tab 31 \tab 24 \tab 25 \tab 81 \tab 26 \tab 144 \tab 51 \tab 81 \tab 46 \tab 660 \tab 57 \tab 125 \tab 33 \tab 271 \tab 24 \tab 62 \tab 19 \tab 56 \tab 189 \tab 459 \tab 5 \tab 2538 \cr
   2016 \tab 43 \tab 8 \tab 11 \tab 40 \tab 26 \tab 24 \tab 80 \tab 32 \tab 157 \tab 60 \tab 74 \tab 65 \tab 774 \tab 72 \tab 147 \tab 35 \tab 274 \tab 28 \tab 67 \tab 16 \tab 59 \tab 221 \tab 550 \tab 4 \tab 2867 \cr
   2018 \tab 42 \tab 1 \tab 12 \tab 33 \tab 15 \tab 25 \tab 64 \tab 33 \tab 112 \tab 47 \tab 54 \tab 50 \tab 596 \tab 55 \tab 136 \tab 29 \tab 238 \tab 19 \tab 65 \tab 13 \tab 48 \tab 194 \tab 464 \tab 3 \tab 2348 \cr
   2021 \tab 42 \tab 2 \tab 13 \tab 34 \tab 18 \tab 40 \tab 82 \tab 41 \tab 177 \tab 71 \tab 102 \tab 69 \tab 1157 \tab 105 \tab 231 \tab 55 \tab 538 \tab 52 \tab 118 \tab 31 \tab 112 \tab 372 \tab 563 \tab 7 \tab 4032 \cr
   2022 \tab 47 \tab 5 \tab 10 \tab 34 \tab 27 \tab 26 \tab 71 \tab 31 \tab 152 \tab 65 \tab 91 \tab 62 \tab 1007 \tab 60 \tab 186 \tab 36 \tab 413 \tab 29 \tab 119 \tab 27 \tab 104 \tab 333 \tab 609 \tab - \tab 3544 \cr
   2024 \tab 34 \tab 2 \tab 8 \tab 23 \tab 16 \tab 26 \tab 69 \tab 26 \tab 134 \tab 51 \tab 72 \tab 49 \tab 967 \tab 92 \tab 179 \tab 53 \tab 355 \tab 33 \tab 82 \tab 20 \tab 98 \tab 303 \tab 617 \tab - \tab 3309 \cr
   Total \tab 1407 \tab 135 \tab 412 \tab 1372 \tab 1190 \tab 1178 \tab 2699 \tab 1337 \tab 6894 \tab 1697 \tab 2341 \tab 1603 \tab 16373 \tab 1593 \tab 3373 \tab 727 \tab 5740 \tab 569 \tab 1469 \tab 429 \tab 1380 \tab 8079 \tab 13351 \tab 351 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{age}},
\code{\link{agekdbrn}},
\code{\link{codeg}},
\code{\link{coeduc}},
\code{\link{degree}},
\code{\link{educ}},
\code{\link{maeduc}},
\code{\link{padeg}},
\code{\link{speduc}}

Other Education: 
\code{\link{degree}},
\code{\link{educ}},
\code{\link{maeduc}},
\code{\link{padeg}},
\code{\link{speduc}}
}
\concept{Core}
\concept{Education}
\keyword{variable}
