% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{paidemps}
\alias{paidemps}
\title{R has paid employees}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidemps
}
\details{
Question Do you have any paid employees?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3818/vshow}{GSS Data Explorer page for \code{paidemps}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 1773 \tab 75 \tab 175 \tab 2023 \cr
   Total \tab 1773 \tab 75 \tab 175 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{owninc}},
\code{\link{whybiz}}
}
\concept{Self Employment}
\keyword{variable}
