% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{paidlv}
\alias{paidlv}
\title{Paid leave for childcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidlv
}
\details{
Question Consider a couple who both work full-time and now have a new born child. One of them stops working for some time to care for their child. Do you think there should be paid leave available and, if so, for how long?
}
\section{Values}{

\itemize{
\item \code{0} no, there should be no paid leave
\item \code{1} yes, enter months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4463/vshow}{GSS Data Explorer page for \code{paidlv}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab yes, enter months \tab 10 \tab 11 \tab 12 \tab 16 \tab 18 \tab no, there should be no paid leave \tab 24 \tab 3 \tab 36 \tab 4 \tab 48 \tab 5 \tab 6 \tab 60 \tab 7 \tab 72 \tab 8 \tab 9 \tab don't know \tab no answer \tab 13 \tab 19 \tab 99 \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 219 \tab 74 \tab 3 \tab 1 \tab 146 \tab 1 \tab 8 \tab 167 \tab 17 \tab 230 \tab 4 \tab 50 \tab 3 \tab 11 \tab 272 \tab 4 \tab 2 \tab 1 \tab 8 \tab 9 \tab 59 \tab 13 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2022 \tab 2383 \tab 87 \tab 57 \tab 3 \tab - \tab 176 \tab 4 \tab 9 \tab 125 \tab 21 \tab 240 \tab 9 \tab 49 \tab 2 \tab 18 \tab 285 \tab 6 \tab 6 \tab - \tab 12 \tab 21 \tab - \tab 21 \tab 1 \tab 1 \tab 1 \tab 7 \tab 3544 \cr
   Total \tab 3055 \tab 306 \tab 131 \tab 6 \tab 1 \tab 322 \tab 5 \tab 17 \tab 292 \tab 38 \tab 470 \tab 13 \tab 99 \tab 5 \tab 29 \tab 557 \tab 10 \tab 8 \tab 1 \tab 20 \tab 30 \tab 59 \tab 34 \tab 1 \tab 1 \tab 1 \tab 7 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
