% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{paidlvdv}
\alias{paidlvdv}
\title{Mother or father paid leave}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidlvdv
}
\details{
Question Still thinking about the same couple, if both are in a similar work situation and are eligible for paid leave, how should this paid leave period be divided between the mother and the father?
}
\section{Values}{

\itemize{
\item \verb{[1]} the mother should take the entire paid leave period and the father should not take any paid leave
\item \verb{[2]} the mother should take most of the paid leave period and the father should take some of it
\item \verb{[3]} the mother and the father should each take half of the paid leave period
\item \verb{[4]} the father should take most of the paid leave period and the mother should take some of it
\item \verb{[5]} the father should take the entire paid leave period and the mother should not take any paid leave
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4466/vshow}{GSS Data Explorer page for \code{paidlvdv}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab the mother should take the entire paid leave period and the father should not take any paid leave \tab the mother should take most of the paid leave period and the father should take some of it \tab the mother and the father should each take half of the paid leave period \tab the father should take most of the paid leave period and the mother should take some of it \tab the father should take the entire paid leave period and the mother should not take any paid leave \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 950 \tab 289 \tab 271 \tab 327 \tab 3 \tab 2 \tab 118 \tab 14 \tab - \tab 1974 \cr
   2022 \tab 2498 \tab 111 \tab 362 \tab 450 \tab 1 \tab 4 \tab 114 \tab - \tab 4 \tab 3544 \cr
   Total \tab 3448 \tab 400 \tab 633 \tab 777 \tab 4 \tab 6 \tab 232 \tab 14 \tab 4 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
