% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{paidlvpy}
\alias{paidlvpy}
\title{Who pays for leave}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidlvpy
}
\details{
Question Consider a couple who both work full-time and now have a new born child. One of them stops working for some time to care for their child. Do you think there should be paid leave available and, if so, for how long?
B. And who should pay for this leave? The government, the employer, both the government and the employer, or other sources?
}
\section{Values}{

\itemize{
\item \code{1} the government
\item \code{2} the employer
\item \code{3} both government and the employer
\item \code{4} other sources
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4465/vshow}{GSS Data Explorer page for \code{paidlvpy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab the government \tab the employer \tab both government and the employer \tab other sources \tab don't know \tab no answer \tab Total \cr
   2012 \tab 950 \tab 69 \tab 359 \tab 489 \tab 53 \tab 41 \tab 13 \tab 1974 \cr
   Total \tab 950 \tab 69 \tab 359 \tab 489 \tab 53 \tab 41 \tab 13 \tab 1974 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
