% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{paidsex}
\alias{paidsex}
\title{R had sex for pay last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidsex
}
\details{
Question If you had other partners, please indicate all categories that apply to them.
D. Person you paid or paid you for sex
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5054/vshow}{GSS Data Explorer page for \code{paidsex}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1988 \tab 1243 \tab 4 \tab 166 \tab 68 \tab - \tab 1481 \cr
   1989 \tab 1303 \tab 4 \tab 156 \tab 74 \tab - \tab 1537 \cr
   1990 \tab 1155 \tab 4 \tab 142 \tab 71 \tab - \tab 1372 \cr
   1991 \tab 1293 \tab 7 \tab 150 \tab 67 \tab - \tab 1517 \cr
   1993 \tab 1359 \tab 4 \tab 156 \tab 87 \tab - \tab 1606 \cr
   1994 \tab 2489 \tab 6 \tab 328 \tab 169 \tab - \tab 2992 \cr
   1996 \tab 2401 \tab 6 \tab 378 \tab 119 \tab - \tab 2904 \cr
   1998 \tab 2383 \tab 12 \tab 283 \tab 154 \tab - \tab 2832 \cr
   2000 \tab 2335 \tab 6 \tab 299 \tab 177 \tab - \tab 2817 \cr
   2002 \tab 2317 \tab 9 \tab 395 \tab 44 \tab - \tab 2765 \cr
   2004 \tab 2429 \tab 10 \tab 367 \tab 6 \tab - \tab 2812 \cr
   2006 \tab 4087 \tab 11 \tab 395 \tab 17 \tab - \tab 4510 \cr
   2008 \tab 1688 \tab 13 \tab 298 \tab 24 \tab - \tab 2023 \cr
   2010 \tab 1705 \tab 13 \tab 296 \tab 30 \tab - \tab 2044 \cr
   2012 \tab 1669 \tab 11 \tab 259 \tab 35 \tab - \tab 1974 \cr
   2014 \tab 2144 \tab 17 \tab 325 \tab 52 \tab - \tab 2538 \cr
   2016 \tab 2530 \tab 11 \tab 306 \tab 20 \tab - \tab 2867 \cr
   2018 \tab 2130 \tab 3 \tab 193 \tab 20 \tab 2 \tab 2348 \cr
   2021 \tab 3747 \tab 11 \tab 243 \tab 20 \tab 11 \tab 4032 \cr
   2022 \tab 3316 \tab 14 \tab 152 \tab 41 \tab 21 \tab 3544 \cr
   Total \tab 43723 \tab 176 \tab 5287 \tab 1295 \tab 34 \tab 50515 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{frndsex}},
\code{\link{matesex}},
\code{\link{othersex}},
\code{\link{partners}},
\code{\link{pikupsex}},
\code{\link{sexfreq}},
\code{\link{sexsex}}
}
\concept{High Risk Behaviors}
\keyword{variable}
