% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{pain}
\alias{pain}
\title{Pain interfere with work past 4 wks}
\source{
General Social Survey https://gss.norc.org
}
\description{
pain
}
\details{
Question During the past 4 weeks, how much did pain interfere with your normal work (including both work outside the home and housework)?
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[2]} a little bit
\item \verb{[3]} moderately
\item \verb{[4]} quite a bit
\item \verb{[5]} extremely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2683/vshow}{GSS Data Explorer page for \code{pain}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab a little bit \tab moderately \tab quite a bit \tab extremely \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 848 \tab 282 \tab 114 \tab 107 \tab 57 \tab 11 \tab 2817 \cr
   Total \tab 1398 \tab 848 \tab 282 \tab 114 \tab 107 \tab 57 \tab 11 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{peaceful}},
\code{\link{socacts}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}
}
\concept{Health Status}
\keyword{variable}
