% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{painarms}
\alias{painarms}
\title{R had pain in the arms in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
painarms
}
\details{
Question B. In the past 12 months, have you had pain in the hands, wrists, arms, or shoulders every day for a week or more?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2832/vshow}{GSS Data Explorer page for \code{painarms}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 969 \tab 504 \tab 1271 \tab 6 \tab 15 \tab - \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 473 \tab 1249 \tab 1 \tab 11 \tab - \tab - \tab 4510 \cr
   2010 \tab 857 \tab 318 \tab 842 \tab 1 \tab 26 \tab - \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 305 \tab 939 \tab 1 \tab 5 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 406 \tab 1008 \tab - \tab 34 \tab - \tab - \tab 2348 \cr
   2022 \tab 1559 \tab 580 \tab 1364 \tab 5 \tab 16 \tab 20 \tab - \tab 3544 \cr
   Total \tab 8349 \tab 2586 \tab 6673 \tab 14 \tab 107 \tab 20 \tab 0 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{chngtme}},
\code{\link{condemnd}},
\code{\link{cowrkhlp}},
\code{\link{cowrkint}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{fairearn}},
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{fringeok}},
\code{\link{handmove}},
\code{\link{haveinfo}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{hlpequip}},
\code{\link{hlthdays}},
\code{\link{hrsrelax}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{hyperten}},
\code{\link{jobfind1}},
\code{\link{jobsecok}},
\code{\link{knowschd}},
\code{\link{knowschd1}},
\code{\link{knowwhat}},
\code{\link{laidoff}},
\code{\link{learnnew}},
\code{\link{lotofsay}},
\code{\link{manvsemp}},
\code{\link{misswork}},
\code{\link{mntlhlth}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{myskills}},
\code{\link{ntwkhard}},
\code{\link{opdevel}},
\code{\link{overwork}},
\code{\link{partteam}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{prodctiv}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{respect}},
\code{\link{rincblls}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{satjob1}},
\code{\link{secondwk}},
\code{\link{setthngs}},
\code{\link{slpprblm}},
\code{\link{spvtrfair}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{suphelp}},
\code{\link{teamsafe}},
\code{\link{toofewwk}},
\code{\link{trainops}},
\code{\link{trdunion}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{usedup}},
\code{\link{usetech}},
\code{\link{waypaid}},
\code{\link{weight}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbonus}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkpraise}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksmooth}},
\code{\link{wktopsat}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrktime}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Health: 
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{anscitst}},
\code{\link{anydoc}},
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{crelesse}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{diffcare}},
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld4}},
\code{\link{disabld5}},
\code{\link{disabld6}},
\code{\link{disabld7}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc15a}},
\code{\link{doc16}},
\code{\link{doc16a}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc18a}},
\code{\link{doc19}},
\code{\link{doc19a}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc20a}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{docaskme}},
\code{\link{doccosts}},
\code{\link{docdecid}},
\code{\link{docinfo}},
\code{\link{doclist}},
\code{\link{docnomri}},
\code{\link{docrely}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{docvig}},
\code{\link{downblue}},
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downfeel}},
\code{\link{downqol}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotfeel}},
\code{\link{emotqol}},
\code{\link{emotrely}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{energy}},
\code{\link{findnwjb}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{health}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{helpsick}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hlthall}},
\code{\link{hlthcare}},
\code{\link{hlthcovr}},
\code{\link{hlthdays}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthinsr}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthplan}},
\code{\link{hlthtv}},
\code{\link{hlthtype}},
\code{\link{hlthwww}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hrdshp6}},
\code{\link{hyperten}},
\code{\link{insrchng}},
\code{\link{insrlmts}},
\code{\link{intmed}},
\code{\link{limitedp}},
\code{\link{misswork}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{modact}},
\code{\link{mykids}},
\code{\link{nodocidu}},
\code{\link{othkids}},
\code{\link{othplan}},
\code{\link{pain}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}},
\code{\link{peaceful}},
\code{\link{physhlth}},
\code{\link{regdoc}},
\code{\link{regdocyr}},
\code{\link{sathealt}},
\code{\link{secopin}},
\code{\link{sexpaid}},
\code{\link{slpprblm}},
\code{\link{socacts}},
\code{\link{sppart}},
\code{\link{stairs}},
\code{\link{swithlth}},
\code{\link{treat1}},
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{usedmntl}},
\code{\link{usedup}},
\code{\link{wait1}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}},
\code{\link{weight}}
}
\concept{Health}
\concept{Quality of Working Life}
\keyword{variable}
