% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_299.R
\name{paincure}
\alias{paincure}
\title{Treatment for pain cure condition}
\source{
General Social Survey https://gss.norc.org
}
\description{
paincure
}
\details{
Question If you went for treatment because you have PAIN that interferes with your work or other daily activities so that you accomplish less than you would like...
C. Would you definitely expect, probably expect, probably not expect, or definitely not expect to be cured?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely expect
\item \verb{[2]} probably expect
\item \verb{[3]} probably not expect
\item \verb{[4]} definitely not expect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2717/vshow}{GSS Data Explorer page for \code{paincure}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely expect \tab probably expect \tab probably not expect \tab definitely not expect \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2339 \tab 122 \tab 246 \tab 58 \tab 16 \tab 19 \tab 17 \tab 2817 \cr
   Total \tab 2339 \tab 122 \tab 246 \tab 58 \tab 16 \tab 19 \tab 17 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downfeel}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotqol}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}}
}
\concept{Health Status}
\keyword{variable}
