% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{paindtag}
\alias{paindtag}
\title{Paind10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
paindtag
}
\details{
Question PAIND10 Coding Tag
}
\section{Values}{

\itemize{
\item \code{1} found
\item \code{2} not found, assigned
\item \code{3} not found, not assigned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5995/vshow}{GSS Data Explorer page for \code{paindtag}}.

Counts by year:\tabular{llllll}{
   year \tab found \tab not found, assigned \tab not found, not assigned \tab iap \tab Total \cr
   1972 \tab 702 \tab 623 \tab 288 \tab - \tab 1613 \cr
   1973 \tab 1234 \tab 159 \tab 111 \tab - \tab 1504 \cr
   1974 \tab 1217 \tab 180 \tab 87 \tab - \tab 1484 \cr
   1975 \tab 976 \tab 395 \tab 119 \tab - \tab 1490 \cr
   1976 \tab 1230 \tab 193 \tab 76 \tab - \tab 1499 \cr
   1977 \tab 817 \tab 527 \tab 186 \tab - \tab 1530 \cr
   1978 \tab 1518 \tab 14 \tab - \tab - \tab 1532 \cr
   1980 \tab 1293 \tab 129 \tab 46 \tab - \tab 1468 \cr
   1982 \tab 1696 \tab 144 \tab 20 \tab - \tab 1860 \cr
   1983 \tab 1475 \tab 101 \tab 23 \tab - \tab 1599 \cr
   1984 \tab 1458 \tab 13 \tab 2 \tab - \tab 1473 \cr
   1985 \tab 1296 \tab 183 \tab 55 \tab - \tab 1534 \cr
   1986 \tab 1425 \tab 35 \tab 10 \tab - \tab 1470 \cr
   1987 \tab 1511 \tab 226 \tab 82 \tab - \tab 1819 \cr
   1988 \tab 1460 \tab 21 \tab - \tab - \tab 1481 \cr
   1989 \tab 1420 \tab 111 \tab 6 \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 14 \tab - \tab - \tab 1372 \cr
   1991 \tab 1499 \tab 17 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 1545 \tab 57 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 2938 \tab 49 \tab 5 \tab - \tab 2992 \cr
   1996 \tab 2883 \tab 19 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 2801 \tab 28 \tab 3 \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   Total \tab 50723 \tab 3238 \tab 1126 \tab 0 \tab 55087 \cr
}
}

\seealso{
Other Core: 
\code{\link{maind80}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{mawrkslf}},
\code{\link{paind10}},
\code{\link{paindstatus}},
\code{\link{papres10}},
\code{\link{papres105plus}}
}
\concept{Core}
\keyword{variable}
