% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{painmed}
\alias{painmed}
\title{Ask doctors to increase pain medication}
\source{
General Social Survey https://gss.norc.org
}
\description{
painmed
}
\details{
Question How much do you agree of disagree with the following options for this person?  This person should...
E. Ask doctors to increase the use of pain medication, even if it might lead to premature death.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2259/vshow}{GSS Data Explorer page for \code{painmed}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 362 \tab 516 \tab 231 \tab 163 \tab 29 \tab 49 \tab 37 \tab 2832 \cr
   Total \tab 1445 \tab 362 \tab 516 \tab 231 \tab 163 \tab 29 \tab 49 \tab 37 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{rightref}},
\code{\link{termvig}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\keyword{variable}
