% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_299.R
\name{painqol}
\alias{painqol}
\title{Treatment for pain improve quality of life}
\source{
General Social Survey https://gss.norc.org
}
\description{
painqol
}
\details{
Question If you went for treatment because you have PAIN that interferes with your work or other daily activities so that you accomplish less than you would like...
A. Would definitely expect, probably expect, probably not expect, or definitely not expect to improve your overall quality of life?
}
\section{Values}{

\itemize{
\item \code{1} definitely expect
\item \code{2} probably expect
\item \code{3} probably not expect
\item \code{4} definitely not expect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2715/vshow}{GSS Data Explorer page for \code{painqol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely expect \tab probably expect \tab probably not expect \tab definitely not expect \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2339 \tab 200 \tab 212 \tab 32 \tab 11 \tab 8 \tab 15 \tab 2817 \cr
   Total \tab 2339 \tab 200 \tab 212 \tab 32 \tab 11 \tab 8 \tab 15 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downfeel}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotqol}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painrely}}
}
\concept{Health Status}
\keyword{variable}
