% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_595.R
\name{paisc681}
\alias{paisc681}
\title{R's father's occupation, 1980 census}
\source{
General Social Survey https://gss.norc.org
}
\description{
paisc681
}
\details{
Question Occupation of respondent's father (International Standard Classification Codes-1968)
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{paisco68}},
\code{\link{spisco68}}
}
\concept{Occupation}
\keyword{variable}
