% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{paisco08}
\alias{paisco08}
\title{R's father's occupation, 2010 census & 2008 isco code}
\source{
General Social Survey https://gss.norc.org
}
\description{
paisco08
}
\details{
Question Respondent's father's occupation, 2010 census, and 2008 ISCO code
}
\section{Values}{

\itemize{
\item \verb{[110]} commissioned armed forces officers
\item \verb{[210]} non-commissioned armed forces officers
\item \verb{[310]} armed forces occupations, other ranks
\item \verb{[1000]} managers
\item \verb{[1100]} chief executives, senior officials and legislators
\item \verb{[1110]} legislators and senior officials
\item \verb{[1111]} legislators
\item \verb{[1112]} senior government officials
\item \verb{[1113]} traditional chiefs and heads of village
\item \verb{[1114]} senior officials of special-interest organizations
\item \verb{[1120]} managing directors and chief executives
\item \verb{[1200]} administrative and commercial managers
\item \verb{[1210]} business services and administration managers
\item \verb{[1211]} finance managers
\item \verb{[1212]} human resource managers
\item \verb{[1213]} policy and planning managers
\item \verb{[1219]} business services and administration managers not elsewhere classified
\item \verb{[1220]} sales, marketing and development managers
\item \verb{[1221]} sales and marketing managers
\item \verb{[1222]} advertising and public relations managers
\item \verb{[1223]} research and development managers
\item \verb{[1300]} production and specialised services managers
\item \verb{[1310]} production managers in agriculture, forestry and fisheries
\item \verb{[1311]} agricultural and forestry production managers
\item \verb{[1312]} aquaculture and fisheries production managers
\item \verb{[1320]} manufacturing, mining, construction, and distribution managers
\item \verb{[1321]} manufacturing managers
\item \verb{[1322]} mining managers
\item \verb{[1323]} construction managers
\item \verb{[1324]} supply, distribution and related managers
\item \verb{[1330]} information and communications technology service managers
\item \verb{[1340]} professional services managers
\item \verb{[1341]} child care services managers
\item \verb{[1342]} health services managers
\item \verb{[1343]} aged care services managers
\item \verb{[1344]} social welfare managers
\item \verb{[1345]} education managers
\item \verb{[1346]} financial and insurance services branch managers
\item \verb{[1349]} professional services managers not elsewhere classified
\item \verb{[1400]} hospitality, retail and other services managers
\item \verb{[1410]} hotel and restaurant managers
\item \verb{[1411]} hotel managers
\item \verb{[1412]} restaurant managers
\item \verb{[1420]} retail and wholesale trade managers
\item \verb{[1430]} other services managers
\item \verb{[1431]} sports, recreation and cultural centre managers
\item \verb{[1439]} services managers not elsewhere classified
\item \verb{[2000]} professionals
\item \verb{[2100]} science and engineering professionals
\item \verb{[2110]} physical and earth science professionals
\item \verb{[2111]} physicists and astronomers
\item \verb{[2112]} meteorologists
\item \verb{[2113]} chemists
\item \verb{[2114]} geologists and geophysicists
\item \verb{[2120]} mathematicians, actuaries and statisticians
\item \verb{[2130]} life science professionals
\item \verb{[2131]} biologists, botanists, zoologists and related professionals
\item \verb{[2132]} farming, forestry and fisheries advisers
\item \verb{[2133]} environmental protection professionals
\item \verb{[2140]} engineering professionals (excluding electrotechnology)
\item \verb{[2141]} industrial and production engineers
\item \verb{[2142]} civil engineers
\item \verb{[2143]} environmental engineers
\item \verb{[2144]} mechanical engineers
\item \verb{[2145]} chemical engineers
\item \verb{[2146]} mining engineers, metallurgists and related professionals
\item \verb{[2149]} engineering professionals not elsewhere classified
\item \verb{[2150]} electrotechnology engineers
\item \verb{[2151]} electrical engineers
\item \verb{[2152]} electronics engineers
\item \verb{[2153]} telecommunications engineers
\item \verb{[2160]} architects, planners, surveyors and designers
\item \verb{[2161]} building architects
\item \verb{[2162]} landscape architects
\item \verb{[2163]} product and garment designers
\item \verb{[2164]} town and traffic planners
\item \verb{[2165]} cartographers and surveyors
\item \verb{[2166]} graphic and multimedia designers
\item \verb{[2200]} health professionals
\item \verb{[2210]} medical doctors
\item \verb{[2211]} generalist medical practitioners
\item \verb{[2212]} specialist medical practitioners
\item \verb{[2220]} nursing and midwifery professionals
\item \verb{[2221]} nursing professionals
\item \verb{[2222]} midwifery professionals
\item \verb{[2230]} traditional and complementary medicine professionals
\item \verb{[2240]} paramedical practitioners
\item \verb{[2250]} veterinarians
\item \verb{[2260]} other health professionals
\item \verb{[2261]} dentists
\item \verb{[2262]} pharmacists
\item \verb{[2263]} environmental and occupational health and hygiene professionals
\item \verb{[2264]} physiotherapists
\item \verb{[2265]} dieticians and nutritionists
\item \verb{[2266]} audiologists and speech therapists
\item \verb{[2267]} optometrists and ophthalmic opticians
\item \verb{[2269]} health professionals not elsewhere classified
\item \verb{[2300]} teaching professionals
\item \verb{[2310]} university and higher education teachers
\item \verb{[2320]} vocational education teachers
\item \verb{[2330]} secondary education teachers
\item \verb{[2340]} primary school and early childhood teachers
\item \verb{[2341]} primary school teachers
\item \verb{[2342]} early childhood educators
\item \verb{[2350]} other teaching professionals
\item \verb{[2351]} education methods specialists
\item \verb{[2352]} special needs teachers
\item \verb{[2353]} other language teachers
\item \verb{[2354]} other music teachers
\item \verb{[2355]} other arts teachers
\item \verb{[2356]} information technology trainers
\item \verb{[2359]} teaching professionals not elsewhere classified
\item \verb{[2400]} business and administration professionals
\item \verb{[2410]} finance professionals
\item \verb{[2411]} accountants
\item \verb{[2412]} financial and investment advisers
\item \verb{[2413]} financial analysts
\item \verb{[2420]} administration professionals
\item \verb{[2421]} management and organization analysts
\item \verb{[2422]} policy administration professionals
\item \verb{[2423]} personnel and careers professionals
\item \verb{[2424]} training and staff development professionals
\item \verb{[2430]} sales, marketing and public relations professionals
\item \verb{[2431]} advertising and marketing professionals
\item \verb{[2432]} public relations professionals
\item \verb{[2433]} technical and medical sales professionals (excluding ict)
\item \verb{[2434]} information and communications technology sales professionals
\item \verb{[2500]} information and communications technology professionals
\item \verb{[2510]} software and applications developers and analysts
\item \verb{[2511]} systems analysts
\item \verb{[2512]} software developers
\item \verb{[2513]} web and multimedia developers
\item \verb{[2514]} applications programmers
\item \verb{[2519]} software and applications developers and analysts not elsewhere classified
\item \verb{[2520]} database and network professionals
\item \verb{[2521]} database designers and administrators
\item \verb{[2522]} systems administrators
\item \verb{[2523]} computer network professionals
\item \verb{[2529]} database and network professionals not elsewhere classified
\item \verb{[2600]} legal, social and cultural professionals
\item \verb{[2610]} legal professionals
\item \verb{[2611]} lawyers
\item \verb{[2612]} judges
\item \verb{[2619]} legal professionals not elsewhere classified
\item \verb{[2620]} librarians, archivists and curators
\item \verb{[2621]} archivists and curators
\item \verb{[2622]} librarians and related information professionals
\item \verb{[2630]} social and religious professionals
\item \verb{[2631]} economists
\item \verb{[2632]} sociologists, anthropologists and related professionals
\item \verb{[2633]} philosophers, historians and political scientists
\item \verb{[2634]} psychologists
\item \verb{[2635]} social work and counselling professionals
\item \verb{[2636]} religious professionals
\item \verb{[2640]} authors, journalists and linguists
\item \verb{[2641]} authors and related writers
\item \verb{[2642]} journalists
\item \verb{[2643]} translators, interpreters and other linguists
\item \verb{[2650]} creative and performing artists
\item \verb{[2651]} visual artists
\item \verb{[2652]} musicians, singers and composers
\item \verb{[2653]} dancers and choreographers
\item \verb{[2654]} film, stage and related directors and producers
\item \verb{[2655]} actors
\item \verb{[2656]} announcers on radio, television and other media
\item \verb{[2659]} creative and performing artists not elsewhere classified
\item \verb{[3000]} technicians and associate professionals
\item \verb{[3100]} science and engineering associate professionals
\item \verb{[3110]} physical and engineering science technicians
\item \verb{[3111]} chemical and physical science technicians
\item \verb{[3112]} civil engineering technicians
\item \verb{[3113]} electrical engineering technicians
\item \verb{[3114]} electronics engineering technicians
\item \verb{[3115]} mechanical engineering technicians
\item \verb{[3116]} chemical engineering technicians
\item \verb{[3117]} mining and metallurgical technicians
\item \verb{[3118]} draughtspersons
\item \verb{[3119]} physical and engineering science technicians not elsewhere classified
\item \verb{[3120]} mining, manufacturing and construction supervisors
\item \verb{[3121]} mining supervisors
\item \verb{[3122]} manufacturing supervisors
\item \verb{[3123]} construction supervisors
\item \verb{[3130]} process control technicians
\item \verb{[3131]} power production plant operators
\item \verb{[3132]} incinerator and water treatment plant operators
\item \verb{[3133]} chemical processing plant controllers
\item \verb{[3134]} petroleum and natural gas refining plant operators
\item \verb{[3135]} metal production process controllers
\item \verb{[3139]} process control technicians not elsewhere classified
\item \verb{[3140]} life science technicians and related associate professionals
\item \verb{[3141]} life science technicians (excluding medical)
\item \verb{[3142]} agricultural technicians
\item \verb{[3143]} forestry technicians
\item \verb{[3150]} ship and aircraft controllers and technicians
\item \verb{[3151]} ships' engineers
\item \verb{[3152]} ships' deck officers and pilots
\item \verb{[3153]} aircraft pilots and related associate professionals
\item \verb{[3154]} air traffic controllers
\item \verb{[3155]} air traffic safety electronics technicians
\item \verb{[3200]} health associate professionals
\item \verb{[3210]} medical and pharmaceutical technicians
\item \verb{[3211]} medical imaging and therapeutic equipment technicians
\item \verb{[3212]} medical and pathology laboratory technicians
\item \verb{[3213]} pharmaceutical technicians and assistants
\item \verb{[3214]} medical and dental prosthetic technicians
\item \verb{[3220]} nursing and midwifery associate professionals
\item \verb{[3221]} nursing associate professionals
\item \verb{[3222]} midwifery associate professionals
\item \verb{[3230]} traditional and complementary medicine associate professionals
\item \verb{[3240]} veterinary technicians and assistants
\item \verb{[3250]} other health associate professionals
\item \verb{[3251]} dental assistants and therapists
\item \verb{[3252]} medical records and health information technicians
\item \verb{[3253]} community health workers
\item \verb{[3254]} dispensing opticians
\item \verb{[3255]} physiotherapy technicians and assistants
\item \verb{[3256]} medical assistants
\item \verb{[3257]} environmental and occupational health inspectors and associates
\item \verb{[3258]} ambulance workers
\item \verb{[3259]} health associate professionals not elsewhere classified
\item \verb{[3300]} business and administration associate professionals
\item \verb{[3310]} financial and mathematical associate professionals
\item \verb{[3311]} securities and finance dealers and brokers
\item \verb{[3312]} credit and loans officers
\item \verb{[3313]} accounting associate professionals
\item \verb{[3314]} statistical, mathematical and related associate professionals
\item \verb{[3315]} valuers and loss assessors
\item \verb{[3320]} sales and purchasing agents and brokers
\item \verb{[3321]} insurance representatives
\item \verb{[3322]} commercial sales representatives
\item \verb{[3323]} buyers
\item \verb{[3324]} trade brokers
\item \verb{[3330]} business services agents
\item \verb{[3331]} clearing and forwarding agents
\item \verb{[3332]} conference and event planners
\item \verb{[3333]} employment agents and contractors
\item \verb{[3334]} real estate agents and property managers
\item \verb{[3339]} business services agents not elsewhere classified
\item \verb{[3340]} administrative and specialised secretaries
\item \verb{[3341]} office supervisors
\item \verb{[3342]} legal secretaries
\item \verb{[3343]} administrative and executive secretaries
\item \verb{[3344]} medical secretaries
\item \verb{[3350]} regulatory government associate professionals
\item \verb{[3351]} customs and border inspectors
\item \verb{[3352]} government tax and excise officials
\item \verb{[3353]} government social benefits officials
\item \verb{[3354]} government licensing officials
\item \verb{[3355]} police inspectors and detectives
\item \verb{[3359]} regulatory government associate professionals not elsewhere classified
\item \verb{[3400]} legal, social, cultural and related associate professionals
\item \verb{[3410]} legal, social and religious associate professionals
\item \verb{[3411]} police inspectors and detectives
\item \verb{[3412]} social work associate professionals
\item \verb{[3413]} religious associate professionals
\item \verb{[3420]} sports and fitness workers
\item \verb{[3421]} athletes and sports players
\item \verb{[3422]} sports coaches, instructors and officials
\item \verb{[3423]} fitness and recreation instructors and program leaders
\item \verb{[3430]} artistic, cultural and culinary associate professionals
\item \verb{[3431]} photographers
\item \verb{[3432]} interior designers and decorators
\item \verb{[3433]} gallery, museum and library technicians
\item \verb{[3434]} chefs
\item \verb{[3435]} other artistic and cultural associate professionals
\item \verb{[3500]} information and communications technicians
\item \verb{[3510]} information and communications technology operations and user support technicians
\item \verb{[3511]} information and communications technology operations technicians
\item \verb{[3512]} information and communications technology user support technicians
\item \verb{[3513]} computer network and systems technicians
\item \verb{[3514]} web technicians
\item \verb{[3520]} telecommunications and broadcasting technicians
\item \verb{[3521]} broadcasting and audio-visual technicians
\item \verb{[3522]} telecommunications engineering technicians
\item \verb{[4000]} clerical support workers
\item \verb{[4100]} general and keyboard clerks
\item \verb{[4110]} general office clerks
\item \verb{[4120]} secretaries (general)
\item \verb{[4130]} keyboard operators
\item \verb{[4131]} typists and word processing operators
\item \verb{[4132]} data entry clerks
\item \verb{[4200]} customer services clerks
\item \verb{[4210]} tellers, money collectors and related clerks
\item \verb{[4211]} bank tellers and related clerks
\item \verb{[4212]} bookmakers, croupiers and related gaming workers
\item \verb{[4213]} pawnbrokers and money-lenders
\item \verb{[4214]} debt-collectors and related workers
\item \verb{[4220]} client information workers
\item \verb{[4221]} travel consultants and clerks
\item \verb{[4222]} contact centre information clerks
\item \verb{[4223]} telephone switchboard operators
\item \verb{[4224]} hotel receptionists
\item \verb{[4225]} enquiry clerks
\item \verb{[4226]} receptionists (general)
\item \verb{[4227]} survey and market research interviewers
\item \verb{[4229]} client information workers not elsewhere classified
\item \verb{[4300]} numerical and material recording clerks
\item \verb{[4310]} numerical clerks
\item \verb{[4311]} accounting and bookkeeping clerks
\item \verb{[4312]} statistical, finance and insurance clerks
\item \verb{[4313]} payroll clerks
\item \verb{[4320]} material-recording and transport clerks
\item \verb{[4321]} stock clerks
\item \verb{[4322]} production clerks
\item \verb{[4323]} transport clerks
\item \verb{[4400]} other clerical support workers
\item \verb{[4410]} other clerical support workers
\item \verb{[4411]} library clerks
\item \verb{[4412]} mail carriers and sorting clerks
\item \verb{[4413]} coding, proof-reading and related clerks
\item \verb{[4414]} scribes and related workers
\item \verb{[4415]} filing and copying clerks
\item \verb{[4416]} personnel clerks
\item \verb{[4419]} clerical support workers not elsewhere classified
\item \verb{[5000]} service and sales workers
\item \verb{[5100]} personal service workers
\item \verb{[5110]} travel attendants, conductors and guides
\item \verb{[5111]} travel attendants and travel stewards
\item \verb{[5112]} transport conductors
\item \verb{[5113]} travel guides
\item \verb{[5120]} cooks
\item \verb{[5130]} waiters and bartenders
\item \verb{[5131]} waiters
\item \verb{[5132]} bartenders
\item \verb{[5140]} hairdressers, beauticians and related workers
\item \verb{[5141]} hairdressers
\item \verb{[5142]} beauticians and related workers
\item \verb{[5150]} building and housekeeping supervisors
\item \verb{[5151]} cleaning and housekeeping supervisors in offices, hotels and other establishments
\item \verb{[5152]} domestic housekeepers
\item \verb{[5153]} building caretakers
\item \verb{[5160]} other personal services workers
\item \verb{[5161]} astrologers, fortune-tellers and related workers
\item \verb{[5162]} companions and valets
\item \verb{[5163]} undertakers and embalmers
\item \verb{[5164]} pet groomers and animal care workers
\item \verb{[5165]} driving instructors
\item \verb{[5169]} personal services workers not elsewhere classified
\item \verb{[5200]} sales workers
\item \verb{[5210]} street and market salespersons
\item \verb{[5211]} stall and market salespersons
\item \verb{[5212]} street food salespersons
\item \verb{[5220]} shop salespersons
\item \verb{[5221]} shop keepers
\item \verb{[5222]} shop supervisors
\item \verb{[5223]} shop sales assistants
\item \verb{[5230]} cashiers and ticket clerks
\item \verb{[5240]} other sales workers
\item \verb{[5241]} fashion and other models
\item \verb{[5242]} sales demonstrators
\item \verb{[5243]} door to door salespersons
\item \verb{[5244]} contact centre salespersons
\item \verb{[5245]} service station attendants
\item \verb{[5246]} food service counter attendants
\item \verb{[5249]} sales workers not elsewhere classified
\item \verb{[5300]} personal care workers
\item \verb{[5310]} child care workers and teachers' aides
\item \verb{[5311]} child care workers
\item \verb{[5312]} teachers' aides
\item \verb{[5320]} personal care workers in health services
\item \verb{[5321]} health care assistants
\item \verb{[5322]} home-based personal care workers
\item \verb{[5329]} personal care workers in health services not elsewhere classified
\item \verb{[5400]} protective services workers
\item \verb{[5410]} protective services workers
\item \verb{[5411]} fire-fighters
\item \verb{[5412]} police officers
\item \verb{[5413]} prison guards
\item \verb{[5414]} security guards
\item \verb{[5419]} protective services workers not elsewhere classified
\item \verb{[6000]} skilled agricultural, forestry and fishery workers
\item \verb{[6100]} market-oriented skilled agricultural workers
\item \verb{[6110]} market gardeners and crop growers
\item \verb{[6111]} field crop and vegetable growers
\item \verb{[6112]} tree and shrub crop growers
\item \verb{[6113]} gardeners, horticultural and nursery growers
\item \verb{[6114]} mixed crop growers
\item \verb{[6120]} animal producers
\item \verb{[6121]} livestock and dairy producers
\item \verb{[6122]} poultry producers
\item \verb{[6123]} apiarists and sericulturists
\item \verb{[6129]} animal producers not elsewhere classified
\item \verb{[6130]} mixed crop and animal producers
\item \verb{[6200]} market-oriented skilled forestry, fishery and hunting workers
\item \verb{[6210]} forestry and related workers
\item \verb{[6220]} fishery workers, hunters and trappers
\item \verb{[6221]} aquaculture workers
\item \verb{[6222]} inland and coastal waters fishery workers
\item \verb{[6223]} deep-sea fishery workers
\item \verb{[6224]} hunters and trappers
\item \verb{[6300]} subsistence farmers, fishers, hunters and gatherers
\item \verb{[6310]} subsistence crop farmers
\item \verb{[6320]} subsistence livestock farmers
\item \verb{[6330]} subsistence mixed crop and livestock farmers
\item \verb{[6340]} subsistence fishers, hunters, trappers and gatherers
\item \verb{[7000]} craft and related trades workers
\item \verb{[7100]} building and related trades workers, excluding electricians
\item \verb{[7110]} building frame and related trades workers
\item \verb{[7111]} house builders
\item \verb{[7112]} bricklayers and related workers
\item \verb{[7113]} stonemasons, stone cutters, splitters and carvers
\item \verb{[7114]} concrete placers, concrete finishers and related workers
\item \verb{[7115]} carpenters and joiners
\item \verb{[7119]} building frame and related trades workers not elsewhere classified
\item \verb{[7120]} building finishers and related trades workers
\item \verb{[7121]} roofers
\item \verb{[7122]} floor layers and tile setters
\item \verb{[7123]} plasterers
\item \verb{[7124]} insulation workers
\item \verb{[7125]} glaziers
\item \verb{[7126]} plumbers and pipe fitters
\item \verb{[7127]} air conditioning and refrigeration mechanics
\item \verb{[7130]} painters, building structure cleaners and related trades workers
\item \verb{[7131]} painters and related workers
\item \verb{[7132]} spray painters and varnishers
\item \verb{[7133]} building structure cleaners
\item \verb{[7200]} metal, machinery and related trades workers
\item \verb{[7210]} sheet and structural metal workers, moulders and welders, and related workers
\item \verb{[7211]} metal moulders and coremakers
\item \verb{[7212]} welders and flamecutters
\item \verb{[7213]} sheet-metal workers
\item \verb{[7214]} structural-metal preparers and erectors
\item \verb{[7215]} riggers and cable splicers
\item \verb{[7220]} blacksmiths, toolmakers and related trades workers
\item \verb{[7221]} blacksmiths, hammersmiths and forging press workers
\item \verb{[7222]} toolmakers and related workers
\item \verb{[7223]} metal working machine tool setters and operators
\item \verb{[7224]} metal polishers, wheel grinders and tool sharpeners
\item \verb{[7230]} machinery mechanics and repairers
\item \verb{[7231]} motor vehicle mechanics and repairers
\item \verb{[7232]} aircraft engine mechanics and repairers
\item \verb{[7233]} agricultural and industrial machinery mechanics and repairers
\item \verb{[7234]} bicycle and related repairers
\item \verb{[7300]} handicraft and printing workers
\item \verb{[7310]} handicraft workers
\item \verb{[7311]} precision-instrument makers and repairers
\item \verb{[7312]} musical instrument makers and tuners
\item \verb{[7313]} jewellery and precious-metal workers
\item \verb{[7314]} potters and related workers
\item \verb{[7315]} glass makers, cutters, grinders and finishers
\item \verb{[7316]} sign writers, decorative painters, engravers and etchers
\item \verb{[7317]} handicraft workers in wood, basketry and related materials
\item \verb{[7318]} handicraft workers in textile, leather and related materials
\item \verb{[7319]} handicraft workers not elsewhere classified
\item \verb{[7320]} printing trades workers
\item \verb{[7321]} pre-press technicians
\item \verb{[7322]} printers
\item \verb{[7323]} print finishing and binding workers
\item \verb{[7400]} electrical and electronic trades workers
\item \verb{[7410]} electrical equipment installers and repairers
\item \verb{[7411]} building and related electricians
\item \verb{[7412]} electrical mechanics and fitters
\item \verb{[7413]} electrical line installers and repairers
\item \verb{[7420]} electronics and telecommunications installers and repairers
\item \verb{[7421]} electronics mechanics and servicers
\item \verb{[7422]} information and communications technology installers and servicers
\item \verb{[7500]} food processing, wood working, garment and other craft and related trades workers
\item \verb{[7510]} food processing and related trades workers
\item \verb{[7511]} butchers, fishmongers and related food preparers
\item \verb{[7512]} bakers, pastry-cooks and confectionery makers
\item \verb{[7513]} dairy-products makers
\item \verb{[7514]} fruit, vegetable and related preservers
\item \verb{[7515]} food and beverage tasters and graders
\item \verb{[7516]} tobacco preparers and tobacco products makers
\item \verb{[7520]} wood treaters, cabinet-makers and related trades workers
\item \verb{[7521]} wood treaters
\item \verb{[7522]} cabinet-makers and related workers
\item \verb{[7523]} woodworking-machine tool setters and operators
\item \verb{[7530]} garment and related trades workers
\item \verb{[7531]} tailors, dressmakers, furriers and hatters
\item \verb{[7532]} garment and related pattern-makers and cutters
\item \verb{[7533]} sewing, embroidery and related workers
\item \verb{[7534]} upholsterers and related workers
\item \verb{[7535]} pelt dressers, tanners and fellmongers
\item \verb{[7536]} shoemakers and related workers
\item \verb{[7540]} other craft and related workers
\item \verb{[7541]} underwater divers
\item \verb{[7542]} shotfirers and blasters
\item \verb{[7543]} product graders and testers (excluding foods and beverages)
\item \verb{[7544]} fumigators and other pest and weed controllers
\item \verb{[7549]} craft and related workers not elsewhere classified
\item \verb{[8000]} plant and machine operators, and assemblers
\item \verb{[8100]} stationary plant and machine operators
\item \verb{[8110]} mining and mineral processing plant operators
\item \verb{[8111]} miners and quarriers
\item \verb{[8112]} mineral and stone processing plant operators
\item \verb{[8113]} well drillers and borers and related workers
\item \verb{[8114]} cement, stone and other mineral products machine operators
\item \verb{[8120]} metal processing and finishing plant operators
\item \verb{[8121]} metal processing plant operators
\item \verb{[8122]} metal finishing, plating and coating machine operators
\item \verb{[8130]} chemical and photographic products plant and machine operators
\item \verb{[8131]} chemical products plant and machine operators
\item \verb{[8132]} photographic products machine operators
\item \verb{[8140]} rubber, plastic and paper products machine operators
\item \verb{[8141]} rubber products machine operators
\item \verb{[8142]} plastic products machine operators
\item \verb{[8143]} paper products machine operators
\item \verb{[8150]} textile, fur and leather products machine operators
\item \verb{[8151]} fibre preparing, spinning and winding machine operators
\item \verb{[8152]} weaving and knitting machine operators
\item \verb{[8153]} sewing machine operators
\item \verb{[8154]} bleaching, dyeing and fabric cleaning machine operators
\item \verb{[8155]} fur and leather preparing machine operators
\item \verb{[8156]} shoemaking and related machine operators
\item \verb{[8157]} laundry machine operators
\item \verb{[8159]} textile, fur and leather products machine operators not elsewhere classified
\item \verb{[8160]} food and related products machine operators
\item \verb{[8170]} wood processing and papermaking plant operators
\item \verb{[8171]} pulp and papermaking plant operators
\item \verb{[8172]} wood processing plant operators
\item \verb{[8180]} other stationary plant and machine operators
\item \verb{[8181]} glass and ceramics plant operators
\item \verb{[8182]} steam engine and boiler operators
\item \verb{[8183]} packing, bottling and labelling machine operators
\item \verb{[8189]} stationary plant and machine operators not elsewhere classified
\item \verb{[8200]} assemblers
\item \verb{[8210]} assemblers
\item \verb{[8211]} mechanical machinery assemblers
\item \verb{[8212]} electrical and electronic equipment assemblers
\item \verb{[8219]} assemblers not elsewhere classified
\item \verb{[8300]} drivers and mobile plant operators
\item \verb{[8310]} locomotive engine drivers and related workers
\item \verb{[8311]} locomotive engine drivers
\item \verb{[8312]} railway brake, signal and switch operators
\item \verb{[8320]} car, van and motorcycle drivers
\item \verb{[8321]} motorcycle drivers
\item \verb{[8322]} car, taxi and van drivers
\item \verb{[8330]} heavy truck and bus drivers
\item \verb{[8331]} bus and tram drivers
\item \verb{[8332]} heavy truck and lorry drivers
\item \verb{[8340]} mobile plant operators
\item \verb{[8341]} mobile farm and forestry plant operators
\item \verb{[8342]} earthmoving and related plant operators
\item \verb{[8343]} crane, hoist and related plant operators
\item \verb{[8344]} lifting truck operators
\item \verb{[8350]} ships' deck crews and related workers
\item \verb{[9000]} elementary occupations
\item \verb{[9100]} cleaners and helpers
\item \verb{[9110]} domestic, hotel and office cleaners and helpers
\item \verb{[9111]} domestic cleaners and helpers
\item \verb{[9112]} cleaners and helpers in offices, hotels and other establishments
\item \verb{[9120]} vehicle, window, laundry and other hand cleaning workers
\item \verb{[9121]} hand launderers and pressers
\item \verb{[9122]} vehicle cleaners
\item \verb{[9123]} window cleaners
\item \verb{[9129]} other cleaning workers
\item \verb{[9200]} agricultural, forestry and fishery labourers
\item \verb{[9210]} agricultural, forestry and fishery labourers
\item \verb{[9211]} crop farm labourers
\item \verb{[9212]} livestock farm labourers
\item \verb{[9213]} mixed crop and livestock farm labourers
\item \verb{[9214]} garden and horticultural labourers
\item \verb{[9215]} forestry labourers
\item \verb{[9216]} fishery and aquaculture labourers
\item \verb{[9300]} labourers in mining, construction, manufacturing and transport
\item \verb{[9310]} mining and construction labourers
\item \verb{[9311]} mining and quarrying labourers
\item \verb{[9312]} civil engineering labourers
\item \verb{[9313]} building construction labourers
\item \verb{[9320]} manufacturing labourers
\item \verb{[9321]} hand packers
\item \verb{[9329]} manufacturing labourers not elsewhere classified
\item \verb{[9330]} transport and storage labourers
\item \verb{[9331]} hand and pedal vehicle drivers
\item \verb{[9332]} drivers of animal-drawn vehicles and machinery
\item \verb{[9333]} freight handlers
\item \verb{[9334]} shelf fillers
\item \verb{[9400]} food preparation assistants
\item \verb{[9410]} food preparation assistants
\item \verb{[9411]} fast food preparers
\item \verb{[9412]} kitchen helpers
\item \verb{[9500]} street and related sales and service workers
\item \verb{[9510]} street and related service workers
\item \verb{[9520]} street vendors (excluding food)
\item \verb{[9600]} refuse workers and other elementary workers
\item \verb{[9610]} refuse workers
\item \verb{[9611]} garbage and recycling collectors
\item \verb{[9612]} refuse sorters
\item \verb{[9613]} sweepers and related labourers
\item \verb{[9620]} other elementary workers
\item \verb{[9621]} messengers, package deliverers and luggage porters
\item \verb{[9622]} odd job persons
\item \verb{[9623]} meter readers and vending-machine collectors
\item \verb{[9624]} water and firewood collectors
\item \verb{[9629]} elementary workers not elsewhere classified
\item \verb{[9997]} uncodeable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{isco08}},
\code{\link{isco88}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{paisco88}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco88}}
}
\concept{Occupation}
\keyword{variable}
