% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{pajwoth}
\alias{pajwoth}
\title{R's dad was jewish when r was growing up}
\source{
General Social Survey https://gss.norc.org
}
\description{
pajwoth
}
\details{
Question Was you father Jewish in any way when you were growing up?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} half/part
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3827/vshow}{GSS Data Explorer page for \code{pajwoth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab half/part \tab no \tab yes \tab not available in this year \tab Total \cr
   2008 \tab 2018 \tab 1 \tab 4 \tab - \tab - \tab 2023 \cr
   2012 \tab 1969 \tab - \tab 4 \tab 1 \tab - \tab 1974 \cr
   Total \tab 3987 \tab 1 \tab 8 \tab 1 \tab 0 \tab 3997 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{bmitzvah}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kidnum}},
\code{\link{majwoth}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{synmem}}

Other Jews: 
\code{\link{bmitzvah}},
\code{\link{comjews}},
\code{\link{fairjews}},
\code{\link{famjews}},
\code{\link{farejews}},
\code{\link{influjew}},
\code{\link{intljews}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{jewtemp}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{knwjew}},
\code{\link{livejews}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{marjew}},
\code{\link{pajew}},
\code{\link{patrjews}},
\code{\link{spjewaj}},
\code{\link{spjoth16}},
\code{\link{usjews}},
\code{\link{violjews}},
\code{\link{wlthjews}},
\code{\link{workjews}}
}
\concept{Jewish Identity}
\concept{Jews}
\keyword{variable}
