% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{palefull}
\alias{palefull}
\title{Afterlife contrast: shadows or full}
\source{
General Social Survey https://gss.norc.org
}
\description{
palefull
}
\details{
Question A. People picture life after death in many different ways. We'd like to know how you think of life after death. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of life after death. Look at the first set of contrasting images. If you think that life after death is "A life like the one here on earth only better," you would place yourself at 1. If you think it is "A life without many things which make our present life enjoyable" you would place yourself at 7. If you think it is somewhere between these two you would place yourself at 2, 3, 4, 5, or 6. Where would you place your images of life after death for . . .?
}
\section{Values}{

\itemize{
\item \code{1} pale, shadowy
\item \code{7} complete fulfillment
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/314/vshow}{GSS Data Explorer page for \code{palefull}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab pale, shadowy \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab complete fulfillment \tab don't know \tab no answer \tab Total \cr
   1986 \tab 246 \tab 42 \tab 15 \tab 39 \tab 170 \tab 110 \tab 137 \tab 674 \tab 13 \tab 24 \tab 1470 \cr
   1987 \tab 369 \tab 41 \tab 14 \tab 26 \tab 184 \tab 163 \tab 231 \tab 714 \tab 22 \tab 55 \tab 1819 \cr
   1988 \tab 282 \tab 37 \tab 21 \tab 32 \tab 210 \tab 143 \tab 168 \tab 536 \tab 15 \tab 37 \tab 1481 \cr
   1989 \tab 749 \tab 21 \tab 13 \tab 19 \tab 142 \tab 113 \tab 98 \tab 334 \tab 17 \tab 31 \tab 1537 \cr
   Total \tab 1646 \tab 141 \tab 63 \tab 116 \tab 706 \tab 529 \tab 634 \tab 2258 \tab 67 \tab 147 \tab 6307 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
