% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{palive}
\alias{palive}
\title{Father still alive?}
\source{
General Social Survey https://gss.norc.org
}
\description{
palive
}
\details{
Question Is your father still alive?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4131/vshow}{GSS Data Explorer page for \code{palive}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1986 \tab - \tab 685 \tab 765 \tab 7 \tab 13 \tab 1470 \cr
   Total \tab 0 \tab 685 \tab 765 \tab 7 \tab 13 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{macall}},
\code{\link{malive}},
\code{\link{matime}},
\code{\link{mavisit}},
\code{\link{peocntct}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
