% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{palive1}
\alias{palive1}
\title{R father alive}
\source{
General Social Survey https://gss.norc.org
}
\description{
palive1
}
\details{
Question Is your father still living?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1457/vshow}{GSS Data Explorer page for \code{palive1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1474 \tab 697 \tab 777 \tab 37 \tab 7 \tab 2992 \cr
   Total \tab 1474 \tab 697 \tab 777 \tab 37 \tab 7 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{malive1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{payrborn}}
}
\concept{Family Mobility}
\keyword{variable}
