% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{pandinc}
\alias{pandinc}
\title{Change in household income over covid pandemic}
\source{
General Social Survey https://gss.norc.org
}
\description{
pandinc
}
\details{
Question Thinking about the income of your household before the COVID-19 pandemic compared with now, would you say it has increased, decreased or stayed about the same?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} increased a lot
\item \verb{[2]} increased a little
\item \verb{[3]} stay about the same
\item \verb{[4]} decreased a little
\item \verb{[5]} decreased a lot
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7954/vshow}{GSS Data Explorer page for \code{pandinc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab stay about the same \tab decreased a little \tab decreased a lot \tab can't choose \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 94 \tab 265 \tab 509 \tab 161 \tab 173 \tab 31 \tab 2 \tab 3544 \cr
   Total \tab 2309 \tab 94 \tab 265 \tab 509 \tab 161 \tab 173 \tab 31 \tab 2 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{fluvax}},
\code{\link{malekids}},
\code{\link{pandmet}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
