% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{pandmet}
\alias{pandmet}
\title{Change in socialization over covid pandemic}
\source{
General Social Survey https://gss.norc.org
}
\description{
pandmet
}
\details{
Question Thinking about how often you met with your extended family and friends in person before the COVID-19 pandemic, would you say you now meet them in person more often, less often, or about the same?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} much less often
\item \code{2} a little less often
\item \code{3} about the same
\item \code{4} a little more often
\item \code{5} much more often
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7955/vshow}{GSS Data Explorer page for \code{pandmet}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much less often \tab a little less often \tab about the same \tab a little more often \tab much more often \tab can't choose \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 202 \tab 322 \tab 549 \tab 84 \tab 62 \tab 15 \tab 1 \tab 3544 \cr
   Total \tab 2309 \tab 202 \tab 322 \tab 549 \tab 84 \tab 62 \tab 15 \tab 1 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{fluvax}},
\code{\link{malekids}},
\code{\link{pandinc}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
