% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{paocc80}
\alias{paocc80}
\title{Father's census occupation code (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
paocc80
}
\details{
Question Father's Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item \verb{[3]} legislators
\item \verb{[4]} chief executives and general admin., public admin
\item \verb{[5]} administrators and officials, public administration
\item \verb{[6]} administrators, protective service
\item \verb{[7]} financial managers
\item \verb{[8]} personnel and labor relations managers
\item \verb{[9]} purchasing managers
\item \verb{[13]} managers, marketing, advertising, and public relations
\item \verb{[14]} administrators, education and related fields
\item \verb{[15]} managers, medicine and health
\item \verb{[16]} managers, properties and real estate
\item \verb{[17]} postmasters and mail superintendents
\item \verb{[18]} funeral directors
\item \verb{[19]} managers and administrators, n.e.c.
\item \verb{[23]} accountants and auditors
\item \verb{[24]} underwriters
\item \verb{[25]} other financial officers
\item \verb{[26]} management analysts
\item \verb{[27]} personnel, training, and labor relations specialists
\item \verb{[28]} purchasing agents and buyers, farm products
\item \verb{[29]} buyers, wholesale and retail trade except farm products
\item \verb{[33]} purchasing agents and buyers
\item \verb{[34]} business and promotion agents
\item \verb{[35]} construction inspectors
\item \verb{[36]} inspectors/compliance officers, except construction
\item \verb{[37]} management related occupations, n.e.c.
\item \verb{[43]} architects
\item \verb{[44]} aerospace engineers
\item \verb{[45]} metallurgical and materials engineers
\item \verb{[46]} mining engineers
\item \verb{[47]} petroleum engineers
\item \verb{[48]} chemical engineers
\item \verb{[49]} nuclear engineers
\item \verb{[53]} civil engineers
\item \verb{[54]} agricultural engineers
\item \verb{[55]} electrical and electronic engineers
\item \verb{[56]} industrial engineers
\item \verb{[57]} mechanical engineers
\item \verb{[58]} marine and naval architects
\item \verb{[59]} engineers, n.e.c.
\item \verb{[63]} surveyors and mapping scientists
\item \verb{[64]} computer systems analysts and scientists
\item \verb{[65]} operations and systems researchers and analysts
\item \verb{[66]} actuaries
\item \verb{[67]} statisticians
\item \verb{[68]} mathematical scientists, n.e.c.
\item \verb{[69]} physicists and astronomers
\item \verb{[73]} chemists, except biochemists
\item \verb{[74]} atmospheric and space scientists
\item \verb{[75]} geologists and geodesists
\item \verb{[76]} physical scientists, n.e.c.
\item \verb{[77]} agricultural and food scientists
\item \verb{[78]} biological and life scientists
\item \verb{[79]} forestry and conservation scientists
\item \verb{[83]} medical scientists
\item \verb{[84]} physicians
\item \verb{[85]} dentists
\item \verb{[86]} veterinarians
\item \verb{[87]} optometrists
\item \verb{[88]} podiatrists
\item \verb{[89]} health diagnosing practitioners, n.e.c.
\item \verb{[95]} registered nurses
\item \verb{[96]} pharmacists
\item \verb{[97]} dietitians
\item \verb{[98]} inhalation therapists
\item \verb{[99]} occupational therapists
\item \verb{[103]} physical therapists
\item \verb{[104]} speech therapists
\item \verb{[105]} therapists, n.e.c.
\item \verb{[106]} physicians' assistants
\item \verb{[113]} earth, environmental, and marine science teachers
\item \verb{[114]} biological science teachers
\item \verb{[115]} chemistry teachers
\item \verb{[116]} physics teachers
\item \verb{[117]} natural science teachers, n.e.c.
\item \verb{[118]} psychology teachers
\item \verb{[119]} economics teachers
\item \verb{[123]} history teachers
\item \verb{[124]} political science teachers
\item \verb{[125]} sociology teachers
\item \verb{[126]} social science teachers, n.e.c.
\item \verb{[127]} engineering teachers
\item \verb{[128]} mathematical science teachers
\item \verb{[129]} computer science teachers
\item \verb{[133]} medical science teachers
\item \verb{[134]} health specialties teachers
\item \verb{[135]} business, commerce, and marketing teachers
\item \verb{[136]} agriculture and forestry teachers
\item \verb{[137]} art, drama, and music teachers
\item \verb{[138]} physical education teachers
\item \verb{[139]} education teachers
\item \verb{[143]} english teachers
\item \verb{[144]} foreign language teachers
\item \verb{[145]} law teachers
\item \verb{[146]} social work teachers
\item \verb{[147]} theology teachers
\item \verb{[148]} trade and industrial teachers
\item \verb{[149]} home economics teachers
\item \verb{[153]} teachers, postsecondary, n.e.c.
\item \verb{[154]} postsecondary teachers, subject not specified
\item \verb{[155]} teachers, prekindergarten and kindergarten
\item \verb{[156]} teachers, elementary school
\item \verb{[157]} teachers, secondary school
\item \verb{[158]} teachers, special education
\item \verb{[159]} teachers, n.e.c.
\item \verb{[163]} counselors, educational and vocational
\item \verb{[164]} librarians
\item \verb{[165]} archivists and curators
\item \verb{[166]} economists
\item \verb{[167]} psychologists
\item \verb{[168]} sociologists
\item \verb{[169]} social scientists, n.e.c.
\item \verb{[173]} urban planners
\item \verb{[174]} social workers
\item \verb{[175]} recreation workers
\item \verb{[176]} clergy
\item \verb{[177]} religious workers, n.e.c.
\item \verb{[178]} lawyers
\item \verb{[179]} judges
\item \verb{[183]} authors
\item \verb{[184]} technical writers
\item \verb{[185]} designers
\item \verb{[186]} musicians and composers
\item \verb{[187]} actors and directors
\item \verb{[188]} painters, sculptors, craft-artists, and printmakers
\item \verb{[189]} photographers
\item \verb{[193]} dancers
\item \verb{[194]} artists, performers, and related workers, n.e.c.
\item \verb{[195]} editors and reporters
\item \verb{[197]} public relations specialists
\item \verb{[198]} announcers
\item \verb{[199]} athletes
\item \verb{[203]} clinical laboratory technologists and technicians
\item \verb{[204]} dental hygienists
\item \verb{[205]} health record technologists and technicians
\item \verb{[206]} radiologic technicians
\item \verb{[207]} licensed practical nurses
\item \verb{[208]} health technologists and technicians, n.e.c.
\item \verb{[213]} electrical and electronic technicians
\item \verb{[214]} industrial engineering technicians
\item \verb{[215]} mechanical engineering technicians
\item \verb{[216]} engineering technicians, n.e.c.
\item \verb{[217]} drafting occupations
\item \verb{[218]} surveying and mapping technicians
\item \verb{[223]} biological technicians
\item \verb{[224]} chemical technicians
\item \verb{[225]} science technicians, n.e.c.
\item \verb{[226]} airplane pilots and navigators
\item \verb{[227]} air traffic controllers
\item \verb{[228]} broadcast equipment operators
\item \verb{[229]} computer programmers
\item \verb{[233]} tool programmers, numerical control
\item \verb{[234]} legal assistants
\item \verb{[235]} technicians, n.e.c.
\item \verb{[243]} supervisors and proprietors, sales occupations
\item \verb{[253]} insurance sales occupations
\item \verb{[254]} real estate sales occupations
\item \verb{[255]} securities and financial services sales occupations
\item \verb{[256]} advertising and related sales occupations
\item \verb{[257]} sales occupations, other business services
\item \verb{[258]} sales engineers
\item \verb{[259]} sales rep., mining, manufacturing, and wholesale
\item \verb{[263]} sales workers, motor vehicles and boats
\item \verb{[264]} sales workers, apparel
\item \verb{[265]} sales workers, shoes
\item \verb{[266]} sales workers, furniture and home furnishings
\item \verb{[267]} sales workers, radio, tv, hi-fi, and appliances
\item \verb{[268]} sales workers, hardware and building supplies
\item \verb{[269]} sales workers, parts
\item \verb{[274]} sales workers, other commodities
\item \verb{[275]} sales counter clerks
\item \verb{[276]} cashiers
\item \verb{[277]} street and door-to-door sales workers
\item \verb{[278]} news vendors
\item \verb{[283]} demonstrators, promoters and models, sales
\item \verb{[284]} auctioneers
\item \verb{[285]} sales support occupations, n.e.c.
\item \verb{[303]} supervisors, general office
\item \verb{[304]} supervisors, computer equipment operators
\item \verb{[305]} supervisors, financial records processing
\item \verb{[306]} chief communications operators
\item \verb{[307]} supervisors, distribution, scheduling, and adjusting clerks
\item \verb{[308]} computer operators
\item \verb{[309]} peripheral equipment operators
\item \verb{[313]} secretaries
\item \verb{[314]} stenographers
\item \verb{[315]} typists
\item \verb{[316]} interviewers
\item \verb{[317]} hotel clerks
\item \verb{[318]} transportation ticket and reservation agents
\item \verb{[319]} receptionists
\item \verb{[323]} information clerks, n.e.c.
\item \verb{[325]} classified-ad clerks
\item \verb{[326]} correspondence clerks
\item \verb{[327]} order clerks
\item \verb{[328]} personnel clerks, except payroll and timekeeping
\item \verb{[329]} library clerks
\item \verb{[335]} file clerks
\item \verb{[336]} records clerks
\item \verb{[337]} bookkeepers, accounting and auditing clerks
\item \verb{[338]} payroll and timekeeping clerks
\item \verb{[339]} billing clerks
\item \verb{[343]} cost and rate clerks
\item \verb{[344]} billing, posting, and calculating machine operators
\item \verb{[345]} duplicating machine operators
\item \verb{[346]} mail preparing and paper handling machine operators
\item \verb{[347]} office machine operators, n.e.c.
\item \verb{[348]} telephone operators
\item \verb{[349]} telegraphers
\item \verb{[353]} communications equipment operators, n.e.c.
\item \verb{[354]} postal clerks, excluding mail carriers
\item \verb{[355]} mail carriers, postal service
\item \verb{[356]} mail clerks, excluding postal service
\item \verb{[357]} messengers
\item \verb{[359]} dispatchers
\item \verb{[363]} production coordinators
\item \verb{[364]} traffic, shipping and receiving clerks
\item \verb{[365]} stock and inventory clerks
\item \verb{[366]} meter readers
\item \verb{[368]} weighers, measurers, and checkers
\item \verb{[369]} samplers
\item \verb{[373]} expediters
\item \verb{[374]} material recording, scheduling and distributing clerks, n.e.c.
\item \verb{[375]} insurance adjusters, examiners, and investigators
\item \verb{[376]} investigators and adjusters, except insurance
\item \verb{[377]} eligibility clerks, social welfare
\item \verb{[378]} bill and account collectors
\item \verb{[379]} general office clerks
\item \verb{[383]} bank tellers
\item \verb{[384]} proofreaders
\item \verb{[385]} data-entry keyers
\item \verb{[386]} statistical clerks
\item \verb{[387]} teachers' aides
\item \verb{[389]} administrative support occupations, n.e.c.
\item \verb{[403]} launderers and ironers
\item \verb{[404]} cooks, private household
\item \verb{[405]} housekeepers and butlers
\item \verb{[406]} child care workers, private household
\item \verb{[407]} private household cleaners and servants
\item \verb{[413]} supervisors, firefighting and fire prevention occupations
\item \verb{[414]} supervisors, police and detectives
\item \verb{[415]} supervisors, guards
\item \verb{[416]} fire inspection and fire prevention occupations
\item \verb{[417]} firefighting occupations
\item \verb{[418]} police and detectives, public service
\item \verb{[423]} sheriffs, bailiffs, and other law enforcement officers
\item \verb{[424]} correctional institution officers
\item \verb{[425]} crossing guards
\item \verb{[426]} guards and police, excluding public service
\item \verb{[427]} protective service occupations
\item \verb{[430]} former member of the armed forces
\item \verb{[431]} current member of the armed forces
\item \verb{[433]} supervisors, food preparation and service occupations
\item \verb{[434]} bartenders
\item \verb{[435]} waiters and waitresses
\item \verb{[436]} cooks, except short order
\item \verb{[437]} short-order cooks
\item \verb{[438]} food counter, fountain and related occupations
\item \verb{[439]} kitchen workers, food preparation
\item \verb{[443]} waiters'/waitresses' assistants
\item \verb{[444]} miscellaneous food preparation occupations
\item \verb{[445]} dental assistants
\item \verb{[446]} health aids, except nursing
\item \verb{[447]} nursing aides, orderlies and attendants
\item \verb{[448]} supervisors, cleaning and building service workers
\item \verb{[449]} maids and housemen
\item \verb{[453]} janitors and cleaners
\item \verb{[454]} elevator operators
\item \verb{[455]} pest control occupations
\item \verb{[456]} supervisors, personal service occupations
\item \verb{[457]} barbers
\item \verb{[458]} hairdressers and cosmetologists
\item \verb{[459]} attendants, amusement and recreation facilities
\item \verb{[463]} guides
\item \verb{[464]} ushers
\item \verb{[465]} public transportation attendants
\item \verb{[466]} baggage porters and bellhops
\item \verb{[467]} welfare service aides
\item \verb{[468]} child care workers, except private household
\item \verb{[469]} personal service occupations, n.e.c.
\item \verb{[473]} farmers, except horticultural
\item \verb{[474]} horticultural specialty farmers
\item \verb{[475]} managers, farms, except horticultural
\item \verb{[476]} managers, horticultural specialty farms
\item \verb{[477]} supervisors, farm workers
\item \verb{[479]} farm workers
\item \verb{[483]} marine life cultivation workers
\item \verb{[484]} nursery workers
\item \verb{[485]} supervisors, related agricultural occupations
\item \verb{[486]} groundskeepers and gardeners, except farm
\item \verb{[487]} animal caretakers, except farm
\item \verb{[488]} graders and sorters, agricultural products
\item \verb{[489]} inspectors, agricultural products
\item \verb{[494]} supervisors, forestry and logging workers
\item \verb{[495]} forestry workers, except logging
\item \verb{[496]} timber cutting and logging occupations
\item \verb{[497]} captains and other officers, fishing vessels
\item \verb{[498]} fishers
\item \verb{[499]} hunters and trappers
\item \verb{[503]} supervisors, mechanics and repairers
\item \verb{[505]} automobile mechanics, except apprentices
\item \verb{[506]} automobile mechanic apprentices
\item \verb{[507]} bus, truck, and stationary engine mechanics
\item \verb{[508]} aircraft engine mechanics
\item \verb{[509]} small engine repairers
\item \verb{[514]} automobile body and related repairers
\item \verb{[515]} aircraft mechanics, excluding engine
\item \verb{[516]} heavy equipment mechanics
\item \verb{[517]} farm equipment mechanics
\item \verb{[518]} industrial machinery repairers
\item \verb{[519]} machinery maintenance occupations
\item \verb{[523]} electronic repairers, communications and industrial equipment
\item \verb{[525]} data processing equipment repairers
\item \verb{[526]} household appliance and power tool repairers
\item \verb{[527]} telephone line installers and repairers
\item \verb{[529]} telephone installers and repairers
\item \verb{[533]} misc. electrical and electronic equipment repairers
\item \verb{[534]} heating, air conditioning, and refrigeration mechanics
\item \verb{[535]} camera, watch, and musical instrument repairers
\item \verb{[536]} locksmiths and safe repairers
\item \verb{[538]} office machine repairers
\item \verb{[539]} mechanical controls and valve repairers
\item \verb{[543]} elevator installers and repairers
\item \verb{[544]} millwrights
\item \verb{[547]} specified mechanics and repairers, n.e.c.
\item \verb{[549]} not specified mechanics and repairers
\item \verb{[553]} supervisors, brickmasons, stonemasons, and title setters
\item \verb{[554]} supervisors, carpenters and related work
\item \verb{[555]} supervisors, electricians and power transmission installers
\item \verb{[556]} supervisors, painters, paperhangers, and plasterers
\item \verb{[557]} supervisors, plumbers, pipefitters, and steamfitters
\item \verb{[558]} supervisors, n.e.c.
\item \verb{[563]} brickmasons and stonemasons, except apprentices
\item \verb{[564]} brickmasons and stonemasons apprentices
\item \verb{[565]} tile setters, hard and soft
\item \verb{[566]} carpet installers
\item \verb{[567]} carpenters, except apprentices
\item \verb{[569]} carpenter apprentices
\item \verb{[573]} drywall installers
\item \verb{[575]} electricians, except apprentices
\item \verb{[576]} electrician apprentices
\item \verb{[577]} electrical power installers and repairers
\item \verb{[579]} painters, construction and maintenance
\item \verb{[583]} paperhangers
\item \verb{[584]} plasterers
\item \verb{[585]} plumbers, pipefitters, and steamfitters, except apprentices
\item \verb{[587]} plumber, pipefitter, and steamfitter apprentices
\item \verb{[588]} concrete and terrazzo finishers
\item \verb{[589]} glaziers
\item \verb{[593]} insulation workers
\item \verb{[594]} paving, surfacing, and tamping equipment operators
\item \verb{[595]} roofers
\item \verb{[596]} sheetmetal duct installers
\item \verb{[597]} structural metal workers
\item \verb{[598]} drillers, earth
\item \verb{[599]} construction trades, n.e.c.
\item \verb{[613]} supervisors, extractive occupations
\item \verb{[614]} drillers, oil well
\item \verb{[615]} explosives workers
\item \verb{[616]} mining machine operators
\item \verb{[617]} mining occupations, n.e.c.
\item \verb{[633]} supervisors, production occupations
\item \verb{[634]} tool and die makers, except apprentices
\item \verb{[635]} tool and die maker apprentices
\item \verb{[636]} precision assemblers, metal
\item \verb{[637]} machinists, except apprentices
\item \verb{[639]} machinist apprentices
\item \verb{[643]} boilermakers
\item \verb{[644]} precision grinders, fitters, and tool sharpeners
\item \verb{[645]} patternmakers and model makers, metal
\item \verb{[646]} lay-out workers
\item \verb{[647]} precious stones and metals workers
\item \verb{[649]} engravers, metal
\item \verb{[653]} sheet metal workers, except apprentices
\item \verb{[654]} sheet metal worker, apprentices
\item \verb{[655]} miscellaneous precision metal workers
\item \verb{[656]} patternmakers and model makers, wood
\item \verb{[657]} cabinet makers and bench carpenters
\item \verb{[658]} furniture and wood finishers
\item \verb{[659]} miscellaneous precision woodworkers
\item \verb{[666]} dressmakers
\item \verb{[667]} tailors
\item \verb{[668]} upholsterers
\item \verb{[669]} shoe repairers
\item \verb{[673]} apparel and fabric patternmakers
\item \verb{[674]} miscellaneous precision apparel and fabric workers
\item \verb{[675]} hand molders and shapers, except jewelers
\item \verb{[676]} patternmakers, lay-out workers, and cutters
\item \verb{[677]} optical goods workers
\item \verb{[678]} dental laboratory and medical appliance technicians
\item \verb{[679]} bookbinders
\item \verb{[683]} electrical and electronic equipment assemblers
\item \verb{[684]} miscellaneous precision workers, n.e.c.
\item \verb{[686]} butchers and meat cutters
\item \verb{[687]} bakers
\item \verb{[688]} food batchmakers
\item \verb{[689]} inspectors, testers, and graders
\item \verb{[693]} adjusters and calibrators
\item \verb{[694]} water and sewage treatment plant operators
\item \verb{[695]} power plant operators
\item \verb{[696]} stationary engineers
\item \verb{[699]} miscellaneous plant and system operators
\item \verb{[703]} lathe and turning machine set-up operators
\item \verb{[704]} lathe and turning machine operators
\item \verb{[705]} milling and planing machine operators
\item \verb{[706]} punching and stamping press machine operators
\item \verb{[707]} rolling machine operators
\item \verb{[708]} drilling and boring machine operators
\item \verb{[709]} grinding, abrading, buffing, and polishing machine operators
\item \verb{[713]} forging machine operators
\item \verb{[714]} numerical control machine operators
\item \verb{[715]} miscellaneous metal, plastic, stone, and glass working machine operators
\item \verb{[717]} fabricating machine operators, n.e.c.
\item \verb{[719]} molding and casting machine operators
\item \verb{[723]} metal plating machine operators
\item \verb{[724]} heat treating equipment operators
\item \verb{[725]} miscellaneous metal and plastic processing machine operators
\item \verb{[726]} wood lathe, routing and planing machine operators
\item \verb{[727]} sawing machine operators
\item \verb{[728]} shaping and joining machine operators
\item \verb{[729]} nailing and tacking machine operators
\item \verb{[733]} miscellaneous woodworking machine operators
\item \verb{[734]} printing machine operators
\item \verb{[735]} photoengravers and lithographers
\item \verb{[736]} typesetters and compositors
\item \verb{[737]} miscellaneous printing machine operators
\item \verb{[738]} winding and twisting machine operators
\item \verb{[739]} knitting, looping, taping, and weaving machine operators
\item \verb{[743]} textile cutting machine operators
\item \verb{[744]} textile sewing machine operators
\item \verb{[745]} shoe machine operators
\item \verb{[747]} pressing machine operators
\item \verb{[748]} laundering and dry cleaning machine operators
\item \verb{[749]} miscellaneous textile machine operators
\item \verb{[753]} cementing and gluing machine operators
\item \verb{[754]} packaging and filling machine operators
\item \verb{[755]} extruding and forming machine operators
\item \verb{[756]} mixing and blending machine operators
\item \verb{[757]} separating, filtering, and clarifying machine operators
\item \verb{[758]} compressing and compacting machine operators
\item \verb{[759]} painting and paint spraying machine operators
\item \verb{[763]} roasting and baking machine operators, food
\item \verb{[764]} washing, cleaning, and pickling machine operators
\item \verb{[765]} folding machine operators
\item \verb{[766]} furnace, kiln, and oven operators, except food
\item \verb{[768]} crushing and grinding machine operators
\item \verb{[769]} slicing and cutting machine operators
\item \verb{[773]} motion picture projectionists
\item \verb{[774]} photographic process machine operators
\item \verb{[777]} miscellaneous and not specified machine operators, n.e.c.
\item \verb{[779]} machine operators, not specified
\item \verb{[783]} welders and cutters
\item \verb{[784]} solderers and blazers
\item \verb{[785]} assemblers
\item \verb{[786]} hand cutting and trimming occupations
\item \verb{[787]} hand molding, casting, and forming occupations
\item \verb{[789]} hand painting,coating, and decorating occupations
\item \verb{[793]} hand engraving and printing occupations
\item \verb{[794]} hand grinding and polishing occupations
\item \verb{[795]} miscellaneous hand working occupations
\item \verb{[796]} production inspectors, checkers, and examiners
\item \verb{[797]} production testers
\item \verb{[798]} production samplers and weighers
\item \verb{[799]} graders and sorters, except agricultural
\item \verb{[803]} supervisors, motor vehicle operators
\item \verb{[804]} truck drivers, heavy
\item \verb{[805]} truck drivers, light
\item \verb{[806]} driver-sales workers
\item \verb{[808]} bus drivers
\item \verb{[809]} taxicab drivers and chauffeurs
\item \verb{[813]} parking lot attendants
\item \verb{[814]} motor transportation occupations, n.e.c.
\item \verb{[823]} railroad conductors and yardmasters
\item \verb{[824]} locomotive operating occupations
\item \verb{[825]} railroad brake, signal, and switch operators
\item \verb{[826]} rail vehicle operators, n.e.c.
\item \verb{[828]} ship captains and mates, except fishing boats
\item \verb{[829]} sailors and deckhands
\item \verb{[833]} marine engineers
\item \verb{[834]} bridge, lock and lighthouse tenders
\item \verb{[843]} supervisors, material moving equipment operators
\item \verb{[844]} operating engineers
\item \verb{[845]} longshore equipment operators
\item \verb{[848]} hoist and winch operators
\item \verb{[849]} crane and tower operators
\item \verb{[853]} excavating and loading machine operators
\item \verb{[855]} grader, dozer, and scraper operators
\item \verb{[856]} industrial truck and tractor equipment operators
\item \verb{[859]} miscellaneous material moving equipment operators
\item \verb{[863]} handlers, equipment cleaners, helpers, and laborers, n.e.c.
\item \verb{[864]} helpers, mechanics and repairers
\item \verb{[865]} helpers, construction trades
\item \verb{[866]} helpers, surveyor
\item \verb{[867]} helpers, extractive occupations
\item \verb{[869]} construction laborers
\item \verb{[873]} production helpers
\item \verb{[875]} garbage collectors
\item \verb{[876]} stevedores
\item \verb{[877]} stock handlers and baggers
\item \verb{[878]} machine feeders and offbearers
\item \verb{[883]} freight, stock, and material handlers, n.e.c.
\item \verb{[885]} garage and service station related occupations
\item \verb{[887]} vehicle washers and equipment cleaners
\item \verb{[888]} hand packers and packagers
\item \verb{[889]} laborers, except construction
\item \verb{[998]} dk
\item \verb{[999]} na
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abhave3}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{ablegal}},
\code{\link{abnomore}},
\code{\link{aborct}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{abspno}},
\code{\link{adoptus}},
\code{\link{adults}},
\code{\link{affrmact}},
\code{\link{age}},
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{agewed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{antiwar}},
\code{\link{arrest}},
\code{\link{attend}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{barate}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blksimp}},
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{bored}},
\code{\link{born}},
\code{\link{brass}},
\code{\link{broadband}},
\code{\link{burglr}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{capfirm}},
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{cathtemp}},
\code{\link{childs}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chlddth}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldsoon}},
\code{\link{cigweek}},
\code{\link{civic}},
\code{\link{civright}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{cobarate}},
\code{\link{codeg}},
\code{\link{coden}},
\code{\link{coden_1222}},
\code{\link{codipged}},
\code{\link{coeduc}},
\code{\link{coeftotlt}},
\code{\link{coevwork}},
\code{\link{cofund}},
\code{\link{cogradtounder}},
\code{\link{cohrs1}},
\code{\link{cohrs2}},
\code{\link{coind10}},
\code{\link{cojail}},
\code{\link{cojew}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{color}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{commute}},
\code{\link{comprend}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlabory}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conorelgsp}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{contemp}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{convictd}},
\code{\link{coocc10}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{coother}},
\code{\link{copres10}},
\code{\link{copunish}},
\code{\link{corel}},
\code{\link{cosector}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{defwrkev}},
\code{\link{defwrknw}},
\code{\link{degree}},
\code{\link{dejavu}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{dotdata}},
\code{\link{dotged}},
\code{\link{dotpeop}},
\code{\link{dotpres}},
\code{\link{dotsvp}},
\code{\link{dotthng}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{earnrs}},
\code{\link{educ}},
\code{\link{educop}},
\code{\link{eftotlt}},
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{era}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{esp}},
\code{\link{evsmoke}},
\code{\link{evwork}},
\code{\link{fair}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{family16sex}},
\code{\link{father}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{febear}},
\code{\link{febrass}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fedirty}},
\code{\link{fefam}},
\code{\link{fefight}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{fehelp}},
\code{\link{fehlpmil}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{femilop}},
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{fenumok}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrite}},
\code{\link{fightair}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{form}},
\code{\link{found}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govcare}},
\code{\link{grace}},
\code{\link{gradtounder}},
\code{\link{granborn}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{guerilla}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}},
\code{\link{gunner}},
\code{\link{gunnum}},
\code{\link{hapcohab}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{healer}},
\code{\link{health}},
\code{\link{helpblk}},
\code{\link{helpful}},
\code{\link{helpnot}},
\code{\link{helpoth}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitnum}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{id}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{immcrime}},
\code{\link{immjobs}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imppol}},
\code{\link{imprelax}},
\code{\link{impwork}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus07}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewtemp}},
\code{\link{jobaccmp}},
\code{\link{jobfind}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{jobtrain}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kidssol}},
\code{\link{king}},
\code{\link{landwar}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{lessnuke}},
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{liberatr}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{libtemp}},
\code{\link{life}},
\code{\link{likediff}},
\code{\link{lockedup}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madeath}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maindstatus}},
\code{\link{maindtag}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{maocctag}},
\code{\link{mapa}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{mapres80}},
\code{\link{marital}},
\code{\link{marsame}},
\code{\link{marsame1}},
\code{\link{marsamey}},
\code{\link{martype}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{mawrkslf}},
\code{\link{mechanic}},
\code{\link{meserve}},
\code{\link{migage}},
\code{\link{mighist}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{mindbody}},
\code{\link{mininc}},
\code{\link{minmilop}},
\code{\link{mobile16}},
\code{\link{mobiledata}},
\code{\link{morenuke}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{neargod}},
\code{\link{news}()},
\code{\link{nonuke}},
\code{\link{norelgsp16}},
\code{\link{nukewar}},
\code{\link{nurse}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{oth16}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{owngun}},
\code{\link{paattend}},
\code{\link{padeath}},
\code{\link{padeg}},
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{paeduc}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{palefull}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres10}},
\code{\link{papres105plus}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{parborn}},
\code{\link{parsol}},
\code{\link{partfull}},
\code{\link{partyid}},
\code{\link{pawrkgrw}},
\code{\link{pawrkslf}},
\code{\link{perla}},
\code{\link{phone}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{pistol}},
\code{\link{polabuse}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{polhitok}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{polmurdr}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{popespks}},
\code{\link{popular}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{premarsx}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prestg10}},
\code{\link{prestg105plus}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{privacy}},
\code{\link{prottemp}},
\code{\link{prowar}},
\code{\link{quitsmk}},
\code{\link{racavoid}},
\code{\link{raccare}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdif5}},
\code{\link{racdif5x}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs15}},
\code{\link{raceacs2}},
\code{\link{raceacs4}},
\code{\link{raceacs6}},
\code{\link{raceacs8}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{racjoin}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{racthink}},
\code{\link{ractrust}},
\code{\link{racwrite}},
\code{\link{radiohrs}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{reg16}},
\code{\link{reg16_7222}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{relgrade}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{res16}},
\code{\link{resdefwk}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{restact}},
\code{\link{richwork}},
\code{\link{rifle}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{robbry}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{rowngun}},
\code{\link{rushed}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfin}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{satjob}},
\code{\link{school}},
\code{\link{sector}},
\code{\link{sex}},
\code{\link{sexeduc}},
\code{\link{sexnow2}},
\code{\link{shotgun}},
\code{\link{sibdeath}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{size}},
\code{\link{smoke}},
\code{\link{smokecig}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{spanking}},
\code{\link{spattend}},
\code{\link{spbarate}},
\code{\link{spdeath}},
\code{\link{spdeg}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spdipged}},
\code{\link{spdotdat}},
\code{\link{spdotged}},
\code{\link{spdotpeo}},
\code{\link{spdotpre}},
\code{\link{spdotsvp}},
\code{\link{spdotthn}},
\code{\link{spdue}},
\code{\link{speduc}},
\code{\link{speftotlt}},
\code{\link{spevwork}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spgradtounder}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spirits}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{spnorelgsp}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sppaid}},
\code{\link{sppres}},
\code{\link{sppres10}},
\code{\link{sppres105plus}},
\code{\link{sppres80}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{spsector}},
\code{\link{spwrkslf}},
\code{\link{spwrksta}},
\code{\link{srcbelt}},
\code{\link{stresswk}},
\code{\link{strike}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{teenpill}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{teensex}},
\code{\link{thnkself}},
\code{\link{ticket}},
\code{\link{tithing}},
\code{\link{transair}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{tvhours}},
\code{\link{typist}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{upgrade}()},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{vetaid}},
\code{\link{vetfam}},
\code{\link{vetfamnw}},
\code{\link{visions}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{weekswrk}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{widowed}},
\code{\link{wirtap}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}},
\code{\link{workhard}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{wrkwayup}},
\code{\link{xmarsex}},
\code{\link{xmovie}},
\code{\link{xmovie1}},
\code{\link{xnorcsiz}},
\code{\link{year}()},
\code{\link{youngen}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{Core}
\concept{Occupation}
\keyword{variable}
