% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{paoth}
\alias{paoth}
\title{Other protestant denomination for father}
\source{
General Social Survey https://gss.norc.org
}
\description{
paoth
}
\details{
Question A. What specific denomination is that, if any?
}
\section{Values}{

\itemize{
\item \code{149} polish catholic
\item \code{196} empowerment temple
\item \code{197} grace independent baptist church
\item \code{198} new life
\item \code{199} follower of christ
\item \code{201} pathways christian church
\item \code{204} renia de deus
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1065/vshow}{GSS Data Explorer page for \code{paoth}}.
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{parelig}},
\code{\link{sunsch16}}

Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
