% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{papernews}
\alias{papernews}
\title{Reliability of newspapers}
\source{
General Social Survey https://gss.norc.org
}
\description{
papernews
}
\details{
Question In general, how reliable would you rate the news on political matters found in the following sources?… Newspapers (in print or access via Internet).
}
\section{Values}{

\itemize{
\item \verb{[1]} very unreliable
\item \verb{[2]} somewhat unreliable
\item \verb{[3]} somewhat reliable
\item \verb{[4]} very reliable
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8163/vshow}{GSS Data Explorer page for \code{papernews}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab very unreliable \tab somewhat unreliable \tab somewhat reliable \tab very reliable \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 147 \tab 288 \tab 746 \tab 207 \tab 236 \tab 1671 \tab 8 \tab 6 \tab 3309 \cr
   Total \tab 0 \tab 147 \tab 288 \tab 746 \tab 207 \tab 236 \tab 1671 \tab 8 \tab 6 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{polnewsfrom}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
