% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{papres80}
\alias{papres80}
\title{Father's occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
papres80
}
\details{
Question Father's Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Core: 
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paocc10}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres16}},
\code{\link{pawrkslf}}

Other Occupation: 
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paocc10}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{pawrkslf}}

Other Occupational Prestige: 
\code{\link{papres16}}
}
\concept{Core}
\concept{Occupation}
\concept{Occupational Prestige}
\keyword{variable}
