% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{parborn}
\alias{parborn}
\title{Were r's parents born in this country}
\source{
General Social Survey https://gss.norc.org
}
\description{
parborn
}
\details{
Question Were both your parents born in this country? (IF THE RESPONDENT IS UNCERTAIN OF PARENT, USE SAME PERSONS AS IN PAOCC16 to PAIND16, PAEDUC, MAEDUC, PADEG, AND MADEG.)
}
\section{Values}{

\itemize{
\item \code{0} both born in the u.s.
\item \code{1} mother yes, father no
\item \code{2} mother no, father yes
\item \code{3} mother yes, father don't know
\item \code{4} mother no, father don't know
\item \code{5} mother don't know, father yes
\item \code{6} mother don't know, father no
\item \code{7} mother don't know, father don't know
\item \code{8} neither born in the u.s.
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/95/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab both born in the u.s. \tab mother don't know, father don't know \tab mother no, father yes \tab mother yes, father don't know \tab mother yes, father no \tab neither born in the u.s. \tab no answer \tab mother no, father don't know \tab mother don't know, father no \tab mother don't know, father yes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1196 \tab 4 \tab 35 \tab 1 \tab 65 \tab 228 \tab 1 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1252 \tab 2 \tab 36 \tab 6 \tab 58 \tab 174 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1178 \tab 3 \tab 30 \tab 3 \tab 58 \tab 193 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1537 \tab 2 \tab 41 \tab 2 \tab 55 \tab 218 \tab 5 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1262 \tab - \tab 44 \tab 1 \tab 59 \tab 229 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1235 \tab 3 \tab 20 \tab - \tab 64 \tab 142 \tab 9 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1265 \tab 2 \tab 34 \tab 3 \tab 58 \tab 168 \tab 4 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1203 \tab - \tab 33 \tab 4 \tab 41 \tab 178 \tab 9 \tab - \tab 1 \tab 1 \tab - \tab 1470 \cr
   1987 \tab - \tab 1559 \tab 3 \tab 37 \tab 2 \tab 45 \tab 161 \tab 10 \tab 2 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 1235 \tab 2 \tab 26 \tab 3 \tab 46 \tab 163 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 1481 \cr
   1989 \tab - \tab 1266 \tab 1 \tab 40 \tab 3 \tab 50 \tab 168 \tab 8 \tab - \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab - \tab 1134 \tab 1 \tab 28 \tab 4 \tab 52 \tab 147 \tab 6 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1253 \tab 3 \tab 40 \tab 2 \tab 64 \tab 144 \tab 9 \tab 1 \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1351 \tab 5 \tab 33 \tab 3 \tab 36 \tab 167 \tab 10 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2480 \tab 3 \tab 71 \tab 8 \tab 91 \tab 318 \tab 17 \tab 2 \tab - \tab 2 \tab - \tab 2992 \cr
   1996 \tab - \tab 2441 \tab 3 \tab 52 \tab 6 \tab 88 \tab 303 \tab 11 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 2323 \tab 2 \tab 47 \tab 11 \tab 86 \tab 338 \tab 20 \tab 1 \tab 3 \tab 1 \tab - \tab 2832 \cr
   2000 \tab - \tab 2306 \tab 4 \tab 75 \tab 5 \tab 86 \tab 332 \tab 6 \tab 3 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2273 \tab - \tab 72 \tab 4 \tab 86 \tab 314 \tab 12 \tab 3 \tab - \tab 1 \tab - \tab 2765 \cr
   2004 \tab - \tab 2295 \tab 1 \tab 61 \tab 2 \tab 80 \tab 359 \tab 10 \tab 3 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 2360 \tab 1 \tab 55 \tab 6 \tab 91 \tab 467 \tab 5 \tab 5 \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab - \tab 1598 \tab - \tab 34 \tab 3 \tab 59 \tab 320 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1631 \tab 1 \tab 44 \tab 2 \tab 40 \tab 314 \tab 10 \tab 1 \tab - \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab 1504 \tab 1 \tab 59 \tab 8 \tab 54 \tab 336 \tab 9 \tab 2 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1950 \tab - \tab 66 \tab 6 \tab 69 \tab 435 \tab 4 \tab 2 \tab 3 \tab 3 \tab - \tab 2538 \cr
   2016 \tab - \tab 2238 \tab 2 \tab 84 \tab 5 \tab 92 \tab 429 \tab 13 \tab 2 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 1823 \tab 4 \tab 62 \tab 7 \tab 72 \tab 371 \tab 6 \tab 3 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 3157 \tab 26 \tab 99 \tab 34 \tab 121 \tab 522 \tab 18 \tab 6 \tab 6 \tab 13 \tab 30 \tab 4032 \cr
   2022 \tab - \tab 2688 \tab 16 \tab 89 \tab 33 \tab 120 \tab 575 \tab 7 \tab 5 \tab 1 \tab 10 \tab - \tab 3544 \cr
   2024 \tab - \tab 2521 \tab 20 \tab 96 \tab 29 \tab 93 \tab 521 \tab 3 \tab 3 \tab 3 \tab 11 \tab 9 \tab 3309 \cr
   Total \tab 9108 \tab 53514 \tab 115 \tab 1543 \tab 206 \tab 2079 \tab 8734 \tab 238 \tab 50 \tab 26 \tab 47 \tab 39 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{born}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{incom16}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}}

Other Nativity: 
\code{\link{born}}
}
\concept{Core}
\concept{Nativity}
\keyword{variable}
