% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{parcit}
\alias{parcit}
\title{Were your parents citizens of america?}
\source{
General Social Survey https://gss.norc.org
}
\description{
parcit
}
\details{
Question B. At the time of your birth, were both, one or neither of your parents citizens of America?
}
\section{Values}{

\itemize{
\item \code{1} both were citizens of america
\item \code{2} only father was a citizen of america
\item \code{3} only mother was a citizen of america
\item \code{4} neither was a citizen of america
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4874/vshow}{GSS Data Explorer page for \code{parcit}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{ethclose}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}}
}
\concept{ISSP National Identity}
\keyword{variable}
