% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{parelig}
\alias{parelig}
\title{What was father's religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
parelig
}
\details{
Question What was your father''s religious preference when you were growing up?  Was it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1063/vshow}{GSS Data Explorer page for \code{parelig}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1988 \tab 202 \tab 703 \tab 360 \tab 28 \tab 151 \tab 30 \tab 1 \tab 6 \tab 1481 \cr
   2008 \tab 1969 \tab 5 \tab 2 \tab 40 \tab 5 \tab - \tab 2 \tab - \tab 2023 \cr
   2012 \tab 1931 \tab 5 \tab 3 \tab 26 \tab 5 \tab 2 \tab 2 \tab - \tab 1974 \cr
   Total \tab 4102 \tab 713 \tab 365 \tab 94 \tab 161 \tab 32 \tab 5 \tab 6 \tab 5478 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
