% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{parelkid}
\alias{parelkid}
\title{Father's religion when r was child}
\source{
General Social Survey https://gss.norc.org
}
\description{
parelkid
}
\details{
Question What was your father''s religious preference when you were a child? Was it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} native american
\item \code{12} other christian
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4667/vshow}{GSS Data Explorer page for \code{parelkid}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab buddhism \tab hinduism \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 704 \tab 368 \tab 32 \tab 9 \tab 4 \tab 14 \tab 89 \tab 114 \tab 25 \tab 1517 \cr
   1998 \tab 1548 \tab 601 \tab 344 \tab 25 \tab 7 \tab 8 \tab 20 \tab 100 \tab 135 \tab 44 \tab 2832 \cr
   2008 \tab 658 \tab 602 \tab 413 \tab 25 \tab 5 \tab 12 \tab 12 \tab 198 \tab 92 \tab 6 \tab 2023 \cr
   2018 \tab 1173 \tab 484 \tab 348 \tab 23 \tab 4 \tab 13 \tab 14 \tab 179 \tab 92 \tab 18 \tab 2348 \cr
   Total \tab 3537 \tab 2391 \tab 1473 \tab 105 \tab 25 \tab 37 \tab 60 \tab 566 \tab 433 \tab 93 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{relexper}},
\code{\link{religkid}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\keyword{variable}
