% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_97.R
\name{parent1}
\alias{parent1}
\title{Person #1 is r's parent}
\source{
General Social Survey https://gss.norc.org
}
\description{
parent1
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
B1. First Person is Parent
}
\section{Values}{

\itemize{
\item \code{1} mentioned
\item \code{2} not mentioned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/880/vshow}{GSS Data Explorer page for \code{parent1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab Total \cr
   1985 \tab 136 \tab 146 \tab 1249 \tab 3 \tab 1534 \cr
   1987 \tab 98 \tab 214 \tab 1487 \tab 20 \tab 1819 \cr
   2004 \tab 1696 \tab 131 \tab 934 \tab 51 \tab 2812 \cr
   Total \tab 1930 \tab 491 \tab 3670 \tab 74 \tab 6165 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}}
}
\concept{Social Networks}
\keyword{variable}
