% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_97.R
\name{parent5}
\alias{parent5}
\title{Person #5 is r's parent}
\source{
General Social Survey https://gss.norc.org
}
\description{
parent5
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
B5. Fifth Person is Parent
}
\section{Values}{

\itemize{
\item \verb{[1]} mentioned
\item \verb{[2]} not mentioned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/884/vshow}{GSS Data Explorer page for \code{parent5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab mentioned \tab not mentioned \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 38 \tab 340 \tab 3 \tab 1534 \cr
   2004 \tab 2600 \tab 11 \tab 155 \tab 46 \tab 2812 \cr
   Total \tab 3753 \tab 49 \tab 495 \tab 49 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}}
}
\concept{Social Networks}
\keyword{variable}
