% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{parlang}
\alias{parlang}
\title{R's parents speak other language at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
parlang
}
\details{
Question Did/do one or both of your parents speak a language other than English at home?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2290/vshow}{GSS Data Explorer page for \code{parlang}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 360 \tab 1013 \tab 2 \tab 23 \tab 2817 \cr
   Total \tab 1419 \tab 360 \tab 1013 \tab 2 \tab 23 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
