% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{parleave}
\alias{parleave}
\title{Offer paid time off to new parents.}
\source{
General Social Survey https://gss.norc.org
}
\description{
parleave
}
\details{
Question It is becoming more common for employers to offer ways for people to combine work and being a parent. For each of the following, please tell me whether or not you think that employers should offer these options to their workers.
B. Offer paid time off to new parents.
}
\section{Values}{

\itemize{
\item \code{1} yes, should offer this option
\item \code{2} no, should not offer this option
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1677/vshow}{GSS Data Explorer page for \code{parleave}}.
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{chldben}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{womenben}}
}
\concept{Gender}
\keyword{variable}
