% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_457.R
\name{parslaw}
\alias{parslaw}
\title{How often does r contact parents-in-law}
\source{
General Social Survey https://gss.norc.org
}
\description{
parslaw
}
\details{
Question Now some questions about your contact with other relatives. Please indicate how often you have been in contact with any of the following types of relatives in the last four weeks.
C. Parents-in-law
}
\section{Values}{

\itemize{
\item \code{1} more than twice in last 4 weeks
\item \code{2} once or twice in last 4 weeks
\item \code{3} not at all in last 4 weeks
\item \code{4} i have no living relative of this type
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4198/vshow}{GSS Data Explorer page for \code{parslaw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab more than twice in last 4 weeks \tab once or twice in last 4 weeks \tab not at all in last 4 weeks \tab i have no living relative of this type \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 223 \tab 141 \tab 221 \tab 555 \tab 2 \tab 7 \tab 2765 \cr
   Total \tab 1616 \tab 223 \tab 141 \tab 221 \tab 555 \tab 2 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{cousins}},
\code{\link{kidcall}},
\code{\link{macall1}},
\code{\link{mavisit1}},
\code{\link{niecenep}},
\code{\link{pacall1}},
\code{\link{pavisit1}},
\code{\link{sibinlaw}},
\code{\link{uncaunts}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
