% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_557.R
\name{partaxes}
\alias{partaxes}
\title{Taxes reduce childbearing}
\source{
General Social Survey https://gss.norc.org
}
\description{
partaxes
}
\details{
Question Please tell me whether you strongly agree, agree, disagree, or strongly disagree with the following statements regarding parents:
I. The burden of taxes discourages couples from having as many children as they would like.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5101/vshow}{GSS Data Explorer page for \code{partaxes}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 82 \tab 458 \tab 563 \tab 132 \tab 117 \tab 20 \tab 1372 \cr
   Total \tab 0 \tab 82 \tab 458 \tab 563 \tab 132 \tab 117 \tab 20 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{rolema}}
}
\concept{Children}
\keyword{variable}
