% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{partopn5}
\alias{partopn5}
\title{Number of sex partner's last 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
partopn5
}
\details{
Question Now, think about the past five years--the time since February/March 1986, and including the past 12 months, how many sex partners have you had in that five year period?
}
\section{Values}{

\itemize{
\item \code{990} -
\item \code{997} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5063/vshow}{GSS Data Explorer page for \code{partopn5}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 14 \tab 15 \tab 19 \tab 2 \tab 20 \tab 25 \tab 3 \tab 4 \tab 40 \tab 5 \tab 50 \tab 6 \tab 65 \tab 7 \tab 8 \tab 80 \tab - \tab don't know \tab no answer \tab Total \cr
   1991 \tab 866 \tab 86 \tab 357 \tab 11 \tab 1 \tab 1 \tab 5 \tab 1 \tab 59 \tab 4 \tab 1 \tab 39 \tab 14 \tab 1 \tab 20 \tab 1 \tab 7 \tab 1 \tab 9 \tab 3 \tab 1 \tab 1 \tab 2 \tab 26 \tab 1517 \cr
   Total \tab 866 \tab 86 \tab 357 \tab 11 \tab 1 \tab 1 \tab 5 \tab 1 \tab 59 \tab 4 \tab 1 \tab 39 \tab 14 \tab 1 \tab 20 \tab 1 \tab 7 \tab 1 \tab 9 \tab 3 \tab 1 \tab 1 \tab 2 \tab 26 \tab 1517 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
