% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{partorg}
\alias{partorg}
\title{In what ways part of the larger org}
\source{
General Social Survey https://gss.norc.org
}
\description{
partorg
}
\details{
Question C. In what way is the place part of the larger organization?  Is it...
}
\section{Values}{

\itemize{
\item \verb{[1]} branch/subsdiary
\item \verb{[2]} org headquarters
\item \verb{[3]} governmnt agency
\item \verb{[4]} public sch sys
\item \verb{[5]} something else
\item \verb{[8]} not sure
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2143/vshow}{GSS Data Explorer page for \code{partorg}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab branch/subsdiary \tab org headquarters \tab governmnt agency \tab public sch sys \tab something else \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2074 \tab 341 \tab 54 \tab 111 \tab 70 \tab 132 \tab 19 \tab 31 \tab 2832 \cr
   Total \tab 2074 \tab 341 \tab 54 \tab 111 \tab 70 \tab 132 \tab 19 \tab 31 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}
}
\concept{Jobs}
\keyword{variable}
