% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_463.R
\name{partpart}
\alias{partpart}
\title{In past 12 months, r has participated in orgs for politics or political associat}
\source{
General Social Survey https://gss.norc.org
}
\description{
partpart
}
\details{
Question In the past 12 months, how often, if at all, have you taken part in the activities? Of political parties, political groups or political associations?
}
\section{Values}{

\itemize{
\item \verb{[1]} once a week or more
\item \verb{[2]} one to three times a month
\item \verb{[3]} several times a year
\item \verb{[4]} once in the past year
\item \verb{[5]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7242/vshow}{GSS Data Explorer page for \code{partpart}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab once a week or more \tab one to three times a month \tab several times a year \tab once in the past year \tab never \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 14 \tab 17 \tab 82 \tab 133 \tab 917 \tab 9 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 14 \tab 17 \tab 82 \tab 133 \tab 917 \tab 9 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
