% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{partpol}
\alias{partpol}
\title{Political participation important for freedom}
\source{
General Social Survey https://gss.norc.org
}
\description{
partpol
}
\details{
Question I'm going to read you several statements about freedom. Each one may say something true and important about freedom, but I'd like to know how important each statement is to you. Is it one of the most important things about freedom, extremely important, very important, somewhat important, or not too important?
D. Freedom is having the right to participate in politics and elections.
}
\section{Values}{

\itemize{
\item \verb{[1]} one of the most important
\item \verb{[2]} extremely important
\item \verb{[3]} very important
\item \verb{[4]} moderately important
\item \verb{[5]} somewhat important
\item \verb{[6]} not too important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2671/vshow}{GSS Data Explorer page for \code{partpol}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab one of the most important \tab extremely important \tab very important \tab moderately important \tab somewhat important \tab not too important \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2000 \tab 1398 \tab 501 \tab 371 \tab 357 \tab 91 \tab 30 \tab 34 \tab 13 \tab 22 \tab - \tab 2817 \cr
   Total \tab 1398 \tab 501 \tab 371 \tab 357 \tab 91 \tab 30 \tab 34 \tab 13 \tab 22 \tab 0 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{freenow}},
\code{\link{freeprss}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{mempolit}},
\code{\link{nogovt}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}},
\code{\link{wlthpov}}

Other Freedom: 
\code{\link{choice}},
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{freenow}},
\code{\link{freeprss}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{mempolit}},
\code{\link{nogovt}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}},
\code{\link{wlthpov}}

Other Leisure: 
\code{\link{artgn}},
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{buyinfgn}},
\code{\link{camping}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{gamegn}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{hrsrelax}},
\code{\link{huntfish}},
\code{\link{investgn}},
\code{\link{makeart}},
\code{\link{memlit}},
\code{\link{memsport}},
\code{\link{newjobgn}},
\code{\link{peoplegn}},
\code{\link{perform}},
\code{\link{perfrmgn}},
\code{\link{plymusic}},
\code{\link{polcangn}},
\code{\link{poldisgn}},
\code{\link{polinfgn}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{seemovie}},
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timepdwk}},
\code{\link{timerelx}},
\code{\link{travelgn}},
\code{\link{usevcr}},
\code{\link{visart}},
\code{\link{visitart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}},
\code{\link{volarts}}

Other Political: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{anomia7}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{bribe}},
\code{\link{buspow}},
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf11}},
\code{\link{caninf12}},
\code{\link{caninf13}},
\code{\link{caninf14}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf6}},
\code{\link{caninf7}},
\code{\link{caninf8}},
\code{\link{caninf9}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{civic}},
\code{\link{cntctgov}},
\code{\link{congeth}},
\code{\link{contemp}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{corruptn}},
\code{\link{demworks}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{engballt}},
\code{\link{ethorgs}},
\code{\link{fixmistk}},
\code{\link{gavepol}},
\code{\link{giveoth}},
\code{\link{govdook}},
\code{\link{governor}},
\code{\link{govngos}},
\code{\link{govtpow}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{interpol}},
\code{\link{intpol}},
\code{\link{joindem}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{leftrght}},
\code{\link{libtemp}},
\code{\link{localgvt}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{ntcitvte}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{owneth}},
\code{\link{partyid}},
\code{\link{polactve}},
\code{\link{polcamp}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff2}},
\code{\link{poleff20}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy3}},
\code{\link{polfrel}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polinfo}},
\code{\link{polint}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polnews}},
\code{\link{polpapr}},
\code{\link{polrally}},
\code{\link{polsfair}},
\code{\link{poltv}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{polwww}},
\code{\link{powerun}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prespop}},
\code{\link{refrndms}},
\code{\link{relcand}},
\code{\link{schlhead}},
\code{\link{servepeo}},
\code{\link{signdpet}},
\code{\link{swayvote}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usrep}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{workpol}}

Other Voluntary Associations: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{bthgrp2a}},
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}},
\code{\link{bthgrp4a}},
\code{\link{bthgrp4b}},
\code{\link{bthgrp4c}},
\code{\link{bthgrp4d}},
\code{\link{bthgrp4e}},
\code{\link{bthgrp5a}},
\code{\link{bthgrp5b}},
\code{\link{bthgrp5c}},
\code{\link{bthgrp5d}},
\code{\link{bthgrp5e}},
\code{\link{caregrp}},
\code{\link{churchtx}},
\code{\link{churhgrp}},
\code{\link{frstgrp1}},
\code{\link{frstgrp2}},
\code{\link{frstgrp3}},
\code{\link{frstgrp4}},
\code{\link{frstgrp5}},
\code{\link{givart}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveduc}},
\code{\link{givegrp}},
\code{\link{givenvir}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{givfound}},
\code{\link{givhlth}},
\code{\link{givhuman}},
\code{\link{givinfrm}},
\code{\link{givintl}},
\code{\link{givoth}},
\code{\link{givpol}},
\code{\link{givpub}},
\code{\link{givrec}},
\code{\link{givrelig}},
\code{\link{givwork}},
\code{\link{givyouth}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grpparty}},
\code{\link{grppol}},
\code{\link{grprelig}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{grpwork}},
\code{\link{hrsart}},
\code{\link{hrseduc}},
\code{\link{hrsenvir}},
\code{\link{hrsfound}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrsinfrm}},
\code{\link{hrsintl}},
\code{\link{hrsoth}},
\code{\link{hrspol}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{hrswork}},
\code{\link{hrsyouth}},
\code{\link{infrmgrp}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{mempolit}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{monart}},
\code{\link{moneduc}},
\code{\link{monenvir}},
\code{\link{monfound}},
\code{\link{monhlth}},
\code{\link{monhuman}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}},
\code{\link{monpol}},
\code{\link{monpub}},
\code{\link{monrec}},
\code{\link{monrelig}},
\code{\link{monwork}},
\code{\link{monyouth}},
\code{\link{mostact}},
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{mtfrat4}},
\code{\link{mtgreek1}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mthobby1}},
\code{\link{mthobby2}},
\code{\link{mthobby3}},
\code{\link{mtinfrm1}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtlit3}},
\code{\link{mtlit4}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}},
\code{\link{mtother1}},
\code{\link{mtother2}},
\code{\link{mtother3}},
\code{\link{mtother4}},
\code{\link{mtother5}},
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}},
\code{\link{mtschl1}},
\code{\link{mtschl2}},
\code{\link{mtschl3}},
\code{\link{mtschl4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv7}},
\code{\link{mtserv8}},
\code{\link{mtsport1}},
\code{\link{mtsport2}},
\code{\link{mtsport3}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtunion4}},
\code{\link{mtvet1}},
\code{\link{mtvet2}},
\code{\link{mtvet3}},
\code{\link{mtvet4}},
\code{\link{mtvet5}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}},
\code{\link{mtyouth4}},
\code{\link{mtyouth5}},
\code{\link{numchurh}},
\code{\link{numfarm}},
\code{\link{numfrat}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numinfrm}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numpolit}},
\code{\link{numprof}},
\code{\link{numschl}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{totart}},
\code{\link{toteduc}},
\code{\link{totenvir}},
\code{\link{totfound}},
\code{\link{tothlth}},
\code{\link{tothuman}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totpub}},
\code{\link{totrec}},
\code{\link{totrelig}},
\code{\link{totwork}},
\code{\link{totyouth}},
\code{\link{typchurh}},
\code{\link{typfarm}},
\code{\link{typfrat}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typinfrm}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typother}},
\code{\link{typpolit}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typserv}},
\code{\link{typsport}},
\code{\link{typunion}},
\code{\link{typvet}},
\code{\link{typyouth}},
\code{\link{valart}},
\code{\link{valeduc}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhlth}},
\code{\link{valhuman}},
\code{\link{valinfrm}},
\code{\link{valintl}},
\code{\link{valoth}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}},
\code{\link{volacty2}},
\code{\link{volactyr}},
\code{\link{volart}},
\code{\link{volarts}},
\code{\link{voleduc}},
\code{\link{volenvir}},
\code{\link{volfound}},
\code{\link{volhlth}},
\code{\link{volhuman}},
\code{\link{volinfrm}},
\code{\link{volintl}},
\code{\link{volmonth}},
\code{\link{voloth}},
\code{\link{volpol}},
\code{\link{volpub}},
\code{\link{volrec}},
\code{\link{volrelig}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{volwork}},
\code{\link{volyouth}},
\code{\link{whoelse1}},
\code{\link{whoelse2}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}},
\code{\link{writegrp}},
\code{\link{yrchurh1}},
\code{\link{yrchurh2}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrfarm3}},
\code{\link{yrfrat1}},
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby1}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrnat1}},
\code{\link{yrnat2}},
\code{\link{yrnat3}},
\code{\link{yrother1}},
\code{\link{yrother2}},
\code{\link{yrother3}},
\code{\link{yrother4}},
\code{\link{yrother5}},
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}},
\code{\link{yrprof6}},
\code{\link{yrschl1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yrserv1}},
\code{\link{yrserv2}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}},
\code{\link{yrserv6}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrsport6}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{Freedom}
\concept{Leisure}
\concept{Political}
\concept{Voluntary Associations}
\keyword{variable}
