% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_307.R
\name{partteam}
\alias{partteam}
\title{R work as part of a team}
\source{
General Social Survey https://gss.norc.org
}
\description{
partteam
}
\details{
Question In your job, do you normally work as part of a team, or do you mostly work on your own?
}
\section{Values}{

\itemize{
\item \code{1} yes, i work as part of a team
\item \code{2} no, i work mostly on my own
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2792/vshow}{GSS Data Explorer page for \code{partteam}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, i work as part of a team \tab no, i work mostly on my own \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 1015 \tab 759 \tab 8 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 850 \tab 861 \tab 11 \tab 12 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 629 \tab 531 \tab 2 \tab 25 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 683 \tab 556 \tab 5 \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 760 \tab 650 \tab 7 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1549 \tab 1172 \tab 788 \tab 6 \tab 13 \tab 16 \tab 3544 \cr
   Total \tab 8339 \tab 5109 \tab 4145 \tab 39 \tab 101 \tab 16 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{prodctiv}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{safehlth}},
\code{\link{setthngs}},
\code{\link{toofewwk}},
\code{\link{trdunion}},
\code{\link{wkdecide}},
\code{\link{wksmooth}}
}
\concept{Quality of Working Life}
\keyword{variable}
