% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{pasei}
\alias{pasei}
\title{R's father's socioeconomic index (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
pasei
}
\details{
Question Socioeconomic index of respondent's father.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} i don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{coisco08}},
\code{\link{firstsei}},
\code{\link{masei}},
\code{\link{pasei10}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{spsei}}
}
\concept{Occupation}
\keyword{variable}
