% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{paslf}
\alias{paslf}
\title{Was father self-employed?}
\source{
General Social Survey https://gss.norc.org
}
\description{
paslf
}
\details{
Question Here is a list of different types of jobs. Which type did your father have when you were 16. (If your father did not have a job then, please give the job he used to have.)
A. Was your father self-employed, or did he work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{3} no dad-never had job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4315/vshow}{GSS Data Explorer page for \code{paslf}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}

Other Occupation: 
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}}
}
\concept{ISSP Social Inequality}
\concept{Occupation}
\keyword{variable}
