% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{pastdown}
\alias{pastdown}
\title{Degree national economy has gotten worse}
\source{
General Social Survey https://gss.norc.org
}
\description{
pastdown
}
\details{
Question How about the economy. Would you say that over the past year the nation's economy has gotten better, stayed about the same, or gotten worse?
B. Would you say much worse or somewhat worse?
}
\section{Values}{

\itemize{
\item \code{1} much worse
\item \code{2} somewhat worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1643/vshow}{GSS Data Explorer page for \code{pastdown}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}}

Other Economy: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{pastup}}
}
\concept{Economy}
\concept{Multiculturalism}
\keyword{variable}
