% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{pastpay}
\alias{pastpay}
\title{Was r's starting pay higher/lower/same as earning in previous job}
\source{
General Social Survey https://gss.norc.org
}
\description{
pastpay
}
\details{
Question When you went to work for your current employer, was your starting pay higher, lower or about the same as what you were earning in your previous (full time) job?
}
\section{Values}{

\itemize{
\item \verb{[1]} higher
\item \verb{[2]} about the same
\item \verb{[3]} lower
\item \verb{[4]} current employer is first job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3875/vshow}{GSS Data Explorer page for \code{pastpay}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab higher \tab about the same \tab lower \tab current employer is first job \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 372 \tab 185 \tab 258 \tab 23 \tab 8 \tab 12 \tab 2023 \cr
   Total \tab 1165 \tab 372 \tab 185 \tab 258 \tab 23 \tab 8 \tab 12 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{curempyr}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{whyleave}},
\code{\link{work3yrs}}
}
\concept{Global Economics}
\keyword{variable}
