% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{patriot3}
\alias{patriot3}
\title{Patriotic feelings are needed for america to remain united}
\source{
General Social Survey https://gss.norc.org
}
\description{
patriot3
}
\details{
Question How much do you agree of disagree that strong patriotic feelings in America…
C. Are needed for America to remain united.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4905/vshow}{GSS Data Explorer page for \code{patriot3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1264 \tab 347 \tab 619 \tab 157 \tab 100 \tab 10 \tab 36 \tab 5 \tab 2538 \cr
   Total \tab 1264 \tab 347 \tab 619 \tab 157 \tab 100 \tab 10 \tab 36 \tab 5 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot4}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\keyword{variable}
