% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{patriot4}
\alias{patriot4}
\title{Patriotic feelings lead to negative feelings towards immigrants}
\source{
General Social Survey https://gss.norc.org
}
\description{
patriot4
}
\details{
Question How much do you agree of disagree that strong patriotic feelings in America…
D. Lead to negative attitudes towards immigrants in America.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4906/vshow}{GSS Data Explorer page for \code{patriot4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1264 \tab 84 \tab 437 \tab 286 \tab 360 \tab 45 \tab 56 \tab 6 \tab 2538 \cr
   Total \tab 1264 \tab 84 \tab 437 \tab 286 \tab 360 \tab 45 \tab 56 \tab 6 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\keyword{variable}
