% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_129.R
\name{patrwhts}
\alias{patrwhts}
\title{Unpatriotic - patriotic}
\source{
General Social Survey https://gss.norc.org
}
\description{
patrwhts
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
F. Do people in these groups tend to be patriotic or do they tend to be unpatriotic?
\enumerate{
\item Where would you rate whites in general on this scale?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} patriotic
\item \verb{[7]} unpatriotic
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1208/vshow}{GSS Data Explorer page for \code{patrwhts}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab patriotic \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab unpatriotic \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 341 \tab 385 \tab 250 \tab 259 \tab 24 \tab 14 \tab 3 \tab 65 \tab 31 \tab 1372 \cr
   Total \tab 0 \tab 341 \tab 385 \tab 250 \tab 259 \tab 24 \tab 14 \tab 3 \tab 65 \tab 31 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{intlhsps}},
\code{\link{intlso}},
\code{\link{patrjews}}
}
\concept{Intergroup Relations}
\keyword{variable}
