% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{pawrkgrw}
\alias{pawrkgrw}
\title{Did father (or oldest same-sex parent) ever work for pay for as long as a year while you were growing up?}
\source{
General Social Survey https://gss.norc.org
}
\description{
pawrkgrw
}
\details{
Question Did father (or oldest same-sex parent) ever work for pay for as long as a year while you were growing up?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8201/vshow}{GSS Data Explorer page for \code{pawrkgrw}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 2497 \tab 105 \tab 102 \tab 597 \tab 5 \tab 3 \tab 3309 \cr
   Total \tab 0 \tab 2497 \tab 105 \tab 102 \tab 597 \tab 5 \tab 3 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conlabory}},
\code{\link{marsamey}},
\code{\link{norelgsp16}},
\code{\link{racdif5}}
}
\concept{Core}
\concept{Family}
\keyword{variable}
