% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_612.R
\name{pawrkslf2}
\alias{pawrkslf2}
\title{At this job, was (dadfill) an employee, self-employed without employees, or self-employed with employees?}
\source{
General Social Survey https://gss.norc.org
}
\description{
pawrkslf2
}
\details{
Question At this job, was your father an employee, self-employed without employees, or self-employed with employees?
}
\section{Values}{

\itemize{
\item \verb{[1]} employee
\item \verb{[2]} self-employed without employees
\item \verb{[3]} self-employed with employees
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7630/vshow}{GSS Data Explorer page for \code{pawrkslf2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab employee \tab self-employed without employees \tab self-employed with employees \tab no answer \tab Total \cr
   2021 \tab 675 \tab 2539 \tab 364 \tab 446 \tab 8 \tab 4032 \cr
   2022 \tab 727 \tab 2197 \tab 303 \tab 314 \tab 3 \tab 3544 \cr
   Total \tab 1402 \tab 4736 \tab 667 \tab 760 \tab 11 \tab 7576 \cr
}
}

\keyword{variable}
