% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{paychnge}
\alias{paychnge}
\title{How much has r pay changed on r's current job since r began}
\source{
General Social Survey https://gss.norc.org
}
\description{
paychnge
}
\details{
Question Thinking about your current employer, how much has your pay changed on your current job since you began?  would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} my pay has gone up more than the cost of living
\item \verb{[2]} my pay has stayed about the same as the cost of living
\item \verb{[3]} my pay has not kept up with the cost of living
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3874/vshow}{GSS Data Explorer page for \code{paychnge}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab my pay has gone up more than the cost of living \tab my pay has stayed about the same as the cost of living \tab my pay has not kept up with the cost of living \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 199 \tab 291 \tab 348 \tab 10 \tab 10 \tab 2023 \cr
   Total \tab 1165 \tab 199 \tab 291 \tab 348 \tab 10 \tab 10 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{curempyr}},
\code{\link{pastpay}},
\code{\link{whyjbct1}},
\code{\link{whyleave}},
\code{\link{work3yrs}}
}
\concept{Global Economics}
\keyword{variable}
