% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{payedtrn}
\alias{payedtrn}
\title{How much pay dependent on educ and training}
\source{
General Social Survey https://gss.norc.org
}
\description{
payedtrn
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
B. The number of year spent in education and training?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4329/vshow}{GSS Data Explorer page for \code{payedtrn}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 263 \tab 566 \tab 319 \tab 41 \tab 10 \tab 23 \tab 50 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 242 \tab 759 \tab 622 \tab 125 \tab 23 \tab 53 \tab 166 \tab 13 \tab 4032 \cr
   Total \tab 3574 \tab 505 \tab 1325 \tab 941 \tab 166 \tab 33 \tab 76 \tab 216 \tab 13 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
