% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_469.R
\name{payexec}
\alias{payexec}
\title{How much r thinks corporate head makes}
\source{
General Social Survey https://gss.norc.org
}
\description{
payexec
}
\details{
Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. nPlease write in how much they actually earn  each year.
E. The chairman of a large national corporation?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4267/vshow}{GSS Data Explorer page for \code{payexec}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10000 \tab 100500 \tab 11000 \tab 110000 \tab 12000 \tab 120000 \tab 125000 \tab 130000 \tab 135000 \tab 15000 \tab 150000 \tab 156000 \tab 16000 \tab 165000 \tab 175000 \tab 18000 \tab 180000 \tab 18200 \tab 184700 \tab 185000 \tab 187000 \tab 189000 \tab 19000 \tab 190000 \tab 197600 \tab 1e+05 \tab $1 million (or more, for 1987 & 2000) \tab 2000 \tab 20000 \tab 21840 \tab 22000 \tab 22500 \tab 225000 \tab 24000 \tab 25000 \tab 250000 \tab 26000 \tab 275000 \tab 28000 \tab 280000 \tab 28800 \tab 29000 \tab 2e+05 \tab 3000 \tab 30000 \tab 31200 \tab 32000 \tab 330000 \tab 34000 \tab 3500 \tab 35000 \tab 350000 \tab 360000 \tab 370000 \tab 375000 \tab 38000 \tab 39000 \tab 3e+05 \tab 4000 \tab 40000 \tab 42000 \tab 420000 \tab 45000 \tab 450000 \tab 47000 \tab 4e+05 \tab 5000 \tab 50000 \tab 55000 \tab 5830 \tab 5e+05 \tab 6000 \tab 60000 \tab 62000 \tab 64000 \tab 65000 \tab 650000 \tab 6e+05 \tab 7000 \tab 70000 \tab 73000 \tab 75000 \tab 750000 \tab 79000 \tab 7e+05 \tab 8000 \tab 80000 \tab 82000 \tab 84000 \tab 8500 \tab 85000 \tab 850000 \tab 89000 \tab 8e+05 \tab 9000 \tab 90000 \tab 93000 \tab 95000 \tab 950000 \tab 9e+05 \tab can't choose \tab no answer \tab 1000 \tab 115000 \tab 12500 \tab 140000 \tab 160000 \tab 170000 \tab 230000 \tab 255000 \tab 270000 \tab 325000 \tab 4500 \tab 49000 \tab 52000 \tab 53000 \tab 57000 \tab 0 \tab 1.1e+07 \tab 1.2e+07 \tab 1.3e+07 \tab 1.4e+07 \tab 1.5e+07 \tab 1.7e+07 \tab 100 \tab 105000 \tab 10500000 \tab 11 \tab 111000 \tab 1200000 \tab 123456 \tab 1250000 \tab 12500000 \tab 1300000 \tab 137000 \tab 1450 \tab 1500000 \tab 150009 \tab 15500000 \tab 1700000 \tab 17200000 \tab 1750000 \tab 176000 \tab 1e+07 \tab 2.5e+07 \tab 20 \tab 210000 \tab 220000 \tab 2200000 \tab 2300000 \tab 240000 \tab 2500000 \tab 25677 \tab 26 \tab 260000 \tab 276000 \tab 2782889 \tab 279000 \tab 290000 \tab 2e+06 \tab 2e+07 \tab 3.5e+07 \tab 30 \tab 320000 \tab 35 \tab 3500000 \tab 380000 \tab 3e+06 \tab 3e+07 \tab 4.5e+07 \tab 4000900 \tab 425000 \tab 4500000 \tab 4566778 \tab 456888 \tab 4e+06 \tab 4e+07 \tab 5 \tab 500 \tab 5500 \tab 550000 \tab 5500000 \tab 5e+06 \tab 5e+07 \tab 600 \tab 676789 \tab 690876 \tab 6e+06 \tab 6e+07 \tab 7.5e+07 \tab 72000 \tab 7500000 \tab 79999 \tab 7e+06 \tab 7e+07 \tab 800 \tab 800009 \tab 8950 \tab 8e+06 \tab 8e+07 \tab 9.9e+07 \tab 98000 \tab 999999 \tab 9e+06 \tab 9e+07 \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 7 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 16 \tab 3 \tab 1 \tab 6 \tab 80 \tab 1 \tab 2 \tab 1 \tab 9 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 202 \tab 66 \tab 2 \tab 13 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 18 \tab 62 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 109 \tab 1 \tab 23 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 13 \tab 11 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 52 \tab 1 \tab 34 \tab 1 \tab 1 \tab 9 \tab 6 \tab 1 \tab 28 \tab 1 \tab 93 \tab 5 \tab 1 \tab 76 \tab 2 \tab 60 \tab 1 \tab 1 \tab 12 \tab 1 \tab 5 \tab 1 \tab 29 \tab 1 \tab 42 \tab 14 \tab 1 \tab 8 \tab 1 \tab 46 \tab 1 \tab 1 \tab 1 \tab 13 \tab 1 \tab 1 \tab 12 \tab 4 \tab 12 \tab 1 \tab 5 \tab 1 \tab 5 \tab 68 \tab 201 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 14 \tab 20 \tab 3 \tab - \tab 1 \tab 56 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 129 \tab 256 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 60 \tab - \tab - \tab - \tab - \tab - \tab - \tab 77 \tab 2 \tab 6 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 7 \tab - \tab - \tab - \tab 1 \tab 1 \tab 35 \tab - \tab 5 \tab - \tab - \tab 5 \tab 3 \tab - \tab 14 \tab - \tab 26 \tab 4 \tab - \tab 96 \tab - \tab 31 \tab - \tab - \tab 2 \tab 1 \tab 10 \tab - \tab 23 \tab - \tab 19 \tab 11 \tab - \tab 5 \tab 1 \tab 36 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 11 \tab - \tab 13 \tab - \tab 2 \tab 1 \tab 2 \tab 75 \tab 164 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 7 \tab - \tab - \tab 1 \tab - \tab 14 \tab 3 \tab 4 \tab - \tab 6 \tab 58 \tab - \tab - \tab - \tab 5 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 55 \tab 238 \tab 1 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 80 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 83 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 15 \tab - \tab - \tab 2 \tab - \tab - \tab 75 \tab - \tab 2 \tab - \tab - \tab - \tab 12 \tab - \tab 51 \tab - \tab 20 \tab 1 \tab - \tab 152 \tab - \tab 5 \tab - \tab - \tab 2 \tab 3 \tab 24 \tab - \tab 4 \tab - \tab 4 \tab 32 \tab - \tab 19 \tab 1 \tab 7 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 19 \tab - \tab 6 \tab - \tab 1 \tab 4 \tab 11 \tab 11 \tab 124 \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 28 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 29 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 127 \tab 22 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 19 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 111 \tab 49 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 67 \tab 14 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 16 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 112 \tab 25 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 50 \tab 4032 \cr
   Total \tab 3829 \tab 17 \tab 1 \tab 1 \tab 8 \tab 5 \tab 34 \tab 39 \tab 10 \tab 1 \tab 13 \tab 194 \tab 1 \tab 2 \tab 1 \tab 16 \tab 2 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 386 \tab 560 \tab 3 \tab 25 \tab 1 \tab 4 \tab 1 \tab 4 \tab 4 \tab 21 \tab 202 \tab 3 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 269 \tab 4 \tab 33 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 17 \tab 33 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab 162 \tab 1 \tab 41 \tab 1 \tab 1 \tab 14 \tab 21 \tab 1 \tab 93 \tab 1 \tab 139 \tab 10 \tab 1 \tab 324 \tab 2 \tab 96 \tab 1 \tab 1 \tab 16 \tab 5 \tab 39 \tab 1 \tab 56 \tab 1 \tab 65 \tab 57 \tab 1 \tab 32 \tab 3 \tab 89 \tab 1 \tab 1 \tab 1 \tab 17 \tab 3 \tab 1 \tab 42 \tab 4 \tab 31 \tab 1 \tab 8 \tab 6 \tab 18 \tab 154 \tab 489 \tab 3 \tab 2 \tab 1 \tab 3 \tab 8 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 28 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 29 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 127 \tab 22 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 19 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 111 \tab 49 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 67 \tab 14 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 16 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 112 \tab 25 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 50 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{inequal7}},
\code{\link{paybus}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payfarm}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{paysec}},
\code{\link{payskill}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
