% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_470.R
\name{payfctry}
\alias{payfctry}
\title{How much r thinks owner of a large factory makes}
\source{
General Social Survey https://gss.norc.org
}
\description{
payfctry
}
\details{
Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. nPlease write in how much they actually earn  each year.
N. The owner-manager of a large factory?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4276/vshow}{GSS Data Explorer page for \code{payfctry}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1000 \tab 10000 \tab 110000 \tab 12000 \tab 120000 \tab 1250 \tab 125000 \tab 13000 \tab 130000 \tab 140000 \tab 1500 \tab 15000 \tab 150000 \tab 160000 \tab 170000 \tab 175000 \tab 180000 \tab 19000 \tab 190000 \tab 1e+05 \tab $1 million (or more, for 1987 & 2000) \tab 20000 \tab 22000 \tab 225000 \tab 23000 \tab 230000 \tab 24000 \tab 25000 \tab 250000 \tab 26000 \tab 28000 \tab 29000 \tab 2e+05 \tab 300 \tab 30000 \tab 32000 \tab 33000 \tab 34000 \tab 35000 \tab 350000 \tab 38000 \tab 39000 \tab 3e+05 \tab 40000 \tab 42000 \tab 45000 \tab 450000 \tab 46000 \tab 4e+05 \tab 50000 \tab 55000 \tab 550000 \tab 56000 \tab 5e+05 \tab 600 \tab 60000 \tab 62400 \tab 64000 \tab 65000 \tab 68000 \tab 69000 \tab 6e+05 \tab 70000 \tab 75000 \tab 750000 \tab 78000 \tab 7e+05 \tab 80000 \tab 85000 \tab 89000 \tab 8e+05 \tab 90000 \tab 95000 \tab 96000 \tab 9e+05 \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 1 \tab 3 \tab 7 \tab 1 \tab 6 \tab 1 \tab 19 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 58 \tab 2 \tab 2 \tab 4 \tab 6 \tab 1 \tab 2 \tab 162 \tab 82 \tab 20 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 7 \tab 44 \tab 1 \tab 1 \tab 1 \tab 65 \tab 1 \tab 23 \tab 1 \tab 1 \tab 1 \tab 23 \tab 2 \tab 2 \tab 1 \tab 39 \tab 33 \tab 1 \tab 15 \tab 2 \tab 1 \tab 8 \tab 66 \tab 4 \tab 1 \tab 1 \tab 64 \tab 1 \tab 48 \tab 1 \tab 1 \tab 11 \tab 2 \tab 1 \tab 4 \tab 27 \tab 37 \tab 5 \tab 1 \tab 3 \tab 49 \tab 5 \tab 1 \tab 6 \tab 25 \tab 2 \tab 1 \tab 2 \tab 83 \tab 152 \tab 2817 \cr
   Total \tab 1545 \tab 1 \tab 3 \tab 7 \tab 1 \tab 6 \tab 1 \tab 19 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 58 \tab 2 \tab 2 \tab 4 \tab 6 \tab 1 \tab 2 \tab 162 \tab 82 \tab 20 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 7 \tab 44 \tab 1 \tab 1 \tab 1 \tab 65 \tab 1 \tab 23 \tab 1 \tab 1 \tab 1 \tab 23 \tab 2 \tab 2 \tab 1 \tab 39 \tab 33 \tab 1 \tab 15 \tab 2 \tab 1 \tab 8 \tab 66 \tab 4 \tab 1 \tab 1 \tab 64 \tab 1 \tab 48 \tab 1 \tab 1 \tab 11 \tab 2 \tab 1 \tab 4 \tab 27 \tab 37 \tab 5 \tab 1 \tab 3 \tab 49 \tab 5 \tab 1 \tab 6 \tab 25 \tab 2 \tab 1 \tab 2 \tab 83 \tab 152 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givmason}},
\code{\link{paycabnt}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{payunskl}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
