% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_469.R
\name{paymason}
\alias{paymason}
\title{How much r thinks bricklayer makes}
\source{
General Social Survey https://gss.norc.org
}
\description{
paymason
}
\details{
Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. nPlease write in how much they actually earn  each year.
A. First, about how much do you think a bricklayer earns?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4263/vshow}{GSS Data Explorer page for \code{paymason}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1000 \tab 10000 \tab 10400 \tab 11000 \tab 11200 \tab 11500 \tab 11520 \tab 12000 \tab 12480 \tab 13000 \tab 14000 \tab 144000 \tab 14560 \tab 15000 \tab 15360 \tab 16000 \tab 16640 \tab 17000 \tab 18000 \tab 18720 \tab 19000 \tab 1e+05 \tab $1 million (or more, for 1987 & 2000) \tab 2000 \tab 20000 \tab 20800 \tab 2100 \tab 21000 \tab 2200 \tab 22000 \tab 22500 \tab 23000 \tab 23040 \tab 24000 \tab 24960 \tab 25000 \tab 250000 \tab 26000 \tab 27000 \tab 27500 \tab 28000 \tab 29000 \tab 29120 \tab 3000 \tab 30000 \tab 30160 \tab 31200 \tab 32000 \tab 32500 \tab 33280 \tab 3400 \tab 3500 \tab 35000 \tab 36000 \tab 37000 \tab 37500 \tab 38000 \tab 39000 \tab 4000 \tab 40000 \tab 40485 \tab 41600 \tab 42000 \tab 42500 \tab 45000 \tab 48000 \tab 5000 \tab 50000 \tab 5200 \tab 52000 \tab 54000 \tab 55000 \tab 6000 \tab 60000 \tab 62400 \tab 65000 \tab 7000 \tab 70000 \tab 75000 \tab 7999 \tab 8000 \tab 80000 \tab 80500 \tab 9000 \tab 90000 \tab 9600 \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 1 \tab 24 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 21 \tab 1 \tab 9 \tab 6 \tab 1 \tab 1 \tab 59 \tab 1 \tab 16 \tab 1 \tab 10 \tab 25 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 181 \tab 3 \tab 1 \tab 3 \tab 1 \tab 19 \tab 1 \tab 7 \tab 1 \tab 19 \tab 2 \tab 180 \tab 1 \tab 5 \tab 12 \tab 1 \tab 14 \tab 6 \tab 1 \tab 5 \tab 272 \tab 1 \tab 4 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab 95 \tab 6 \tab 2 \tab 1 \tab 9 \tab 1 \tab 2 \tab 115 \tab 1 \tab 1 \tab 1 \tab 1 \tab 22 \tab 1 \tab 3 \tab 45 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 16 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 68 \tab 187 \tab 1819 \cr
   Total \tab 255 \tab 1 \tab 24 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 21 \tab 1 \tab 9 \tab 6 \tab 1 \tab 1 \tab 59 \tab 1 \tab 16 \tab 1 \tab 10 \tab 25 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 181 \tab 3 \tab 1 \tab 3 \tab 1 \tab 19 \tab 1 \tab 7 \tab 1 \tab 19 \tab 2 \tab 180 \tab 1 \tab 5 \tab 12 \tab 1 \tab 14 \tab 6 \tab 1 \tab 5 \tab 272 \tab 1 \tab 4 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab 95 \tab 6 \tab 2 \tab 1 \tab 9 \tab 1 \tab 2 \tab 115 \tab 1 \tab 1 \tab 1 \tab 1 \tab 22 \tab 1 \tab 3 \tab 45 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 16 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 68 \tab 187 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{inequal7}},
\code{\link{paybus}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payowner}},
\code{\link{paysec}},
\code{\link{payskill}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
