% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{payresp}
\alias{payresp}
\title{How much pay dependent on responsibility}
\source{
General Social Survey https://gss.norc.org
}
\description{
payresp
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
A How much responsibility goes with the job"how important do you think that ought to be in deciding pay?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4328/vshow}{GSS Data Explorer page for \code{payresp}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 348 \tab 631 \tab 192 \tab 14 \tab 3 \tab 34 \tab 50 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 436 \tab 950 \tab 357 \tab 12 \tab 12 \tab 61 \tab 164 \tab 11 \tab 4032 \cr
   Total \tab 3574 \tab 784 \tab 1581 \tab 549 \tab 26 \tab 15 \tab 95 \tab 214 \tab 11 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
