% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{paysame}
\alias{paysame}
\title{Importance of standard rates to pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
paysame
}
\details{
Question Think of two people doing the same kind of work. What do you personally think should be important in deciding how much to pay them?
Looking at the things below, please write ''1'' next to the thing that you think should be most important. Then write ''2'' next to the thing that you think should be next most important, and ''3'' next to the thing that you think should be third most important. Leave the other boxes blank.
D. The standard rate, giving both employees the same pay
}
\section{Values}{

\itemize{
\item \verb{[1]} most impt
\item \verb{[2]} 2nd most impt
\item \verb{[3]} 3rd most impt
\item \verb{[4]} not chosen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4499/vshow}{GSS Data Explorer page for \code{paysame}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab most impt \tab 2nd most impt \tab 3rd most impt \tab not chosen \tab don't know \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 44 \tab 50 \tab 76 \tab 921 \tab 53 \tab 393 \tab - \tab 1537 \cr
   Total \tab 0 \tab 44 \tab 50 \tab 76 \tab 921 \tab 53 \tab 393 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{hwduties}},
\code{\link{sexemp}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}},
\code{\link{yrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
