% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_469.R
\name{paysec}
\alias{paysec}
\title{How much r thinks secretary makes}
\source{
General Social Survey https://gss.norc.org
}
\description{
paysec
}
\details{
Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. nPlease write in how much they actually earn  each year.
H. A secretary?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4270/vshow}{GSS Data Explorer page for \code{paysec}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1000 \tab 10000 \tab 10400 \tab 10500 \tab 11000 \tab 12000 \tab 12499 \tab 12500 \tab 13000 \tab 13440 \tab 13500 \tab 14000 \tab 14400 \tab 14560 \tab 1500 \tab 15000 \tab 15600 \tab 16000 \tab 16640 \tab 17000 \tab 17250 \tab 17500 \tab 17680 \tab 18000 \tab 18500 \tab 18720 \tab 19000 \tab $1 million (or more, for 1987 & 2000) \tab 2000 \tab 20000 \tab 20500 \tab 21000 \tab 21120 \tab 22000 \tab 22500 \tab 23000 \tab 230000 \tab 23500 \tab 24000 \tab 25000 \tab 26000 \tab 27000 \tab 28000 \tab 3000 \tab 30000 \tab 32000 \tab 3250 \tab 35000 \tab 36000 \tab 36400 \tab 4000 \tab 40000 \tab 45000 \tab 4800 \tab 5000 \tab 50000 \tab 5200 \tab 6000 \tab 60000 \tab 61000 \tab 62400 \tab 65000 \tab 7000 \tab 7500 \tab 7800 \tab 8000 \tab 8500 \tab 9000 \tab 9100 \tab 9600 \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 2 \tab 90 \tab 4 \tab 1 \tab 17 \tab 148 \tab 1 \tab 8 \tab 46 \tab 1 \tab 4 \tab 52 \tab 1 \tab 1 \tab 1 \tab 242 \tab 1 \tab 58 \tab 1 \tab 40 \tab 1 \tab 5 \tab 1 \tab 130 \tab 3 \tab 1 \tab 17 \tab 1 \tab 7 \tab 162 \tab 1 \tab 12 \tab 1 \tab 19 \tab 1 \tab 10 \tab 1 \tab 1 \tab 6 \tab 72 \tab 1 \tab 2 \tab 7 \tab 1 \tab 23 \tab 1 \tab 1 \tab 10 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 9 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 3 \tab 1 \tab 30 \tab 3 \tab 18 \tab 1 \tab 3 \tab 63 \tab 186 \tab 1819 \cr
   Total \tab 255 \tab 2 \tab 90 \tab 4 \tab 1 \tab 17 \tab 148 \tab 1 \tab 8 \tab 46 \tab 1 \tab 4 \tab 52 \tab 1 \tab 1 \tab 1 \tab 242 \tab 1 \tab 58 \tab 1 \tab 40 \tab 1 \tab 5 \tab 1 \tab 130 \tab 3 \tab 1 \tab 17 \tab 1 \tab 7 \tab 162 \tab 1 \tab 12 \tab 1 \tab 19 \tab 1 \tab 10 \tab 1 \tab 1 \tab 6 \tab 72 \tab 1 \tab 2 \tab 7 \tab 1 \tab 23 \tab 1 \tab 1 \tab 10 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 9 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 3 \tab 1 \tab 30 \tab 3 \tab 18 \tab 1 \tab 3 \tab 63 \tab 186 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{inequal7}},
\code{\link{paybus}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payskill}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
