% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{paysup}
\alias{paysup}
\title{How much pay dependent on supervising others}
\source{
General Social Survey https://gss.norc.org
}
\description{
paysup
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
C. Whether the job requires supervising others (how important should that be in deciding pay)?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4330/vshow}{GSS Data Explorer page for \code{paysup}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 183 \tab 572 \tab 357 \tab 62 \tab 10 \tab 39 \tab 49 \tab 2817 \cr
   Total \tab 1545 \tab 183 \tab 572 \tab 357 \tab 62 \tab 10 \tab 39 \tab 49 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
