% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{paytime}
\alias{paytime}
\title{How important be long with the firm in deciding p}
\source{
General Social Survey https://gss.norc.org
}
\description{
paytime
}
\details{
Question In deciding on pay for two people doing the same kind of work, how important should be …
D. How long the person has been with the firm?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4574/vshow}{GSS Data Explorer page for \code{paytime}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1609 \tab 141 \tab 418 \tab 407 \tab 144 \tab 50 \tab 23 \tab 40 \tab 2832 \cr
   Total \tab 1609 \tab 141 \tab 418 \tab 407 \tab 144 \tab 50 \tab 23 \tab 40 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{wantjob1}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
