% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{paywww}
\alias{paywww}
\title{Does employer pay for www}
\source{
General Social Survey https://gss.norc.org
}
\description{
paywww
}
\details{
Question B. Does your employer provide all or part of your Internet access from home?
}
\section{Values}{

\itemize{
\item \verb{[1]} employer provides internet connection at no charge
\item \verb{[2]} covers some but not all of cost
\item \verb{[3]} does not contribute to cost of using internet
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2553/vshow}{GSS Data Explorer page for \code{paywww}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab employer provides internet connection at no charge \tab covers some but not all of cost \tab does not contribute to cost of using internet \tab no answer \tab Total \cr
   2000 \tab 2471 \tab 42 \tab 8 \tab 243 \tab 53 \tab 2817 \cr
   Total \tab 2471 \tab 42 \tab 8 \tab 243 \tab 53 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
