% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{peocntct}
\alias{peocntct}
\title{How many people in contact in a typical weekday}
\source{
General Social Survey https://gss.norc.org
}
\description{
peocntct
}
\details{
Question On average, about how many people do you have contact with in a typical week day, including people you live with.
We are interested in contact on a  one-to-one basis, including everyone with whom you chat, talk, or discuss matters.
This can be face-to-face, by telephone, by mail, or on the internet. Please include only people you know.
Please tell me which of the following categories best matches your estimate.
Would you say…
}
\section{Values}{

\itemize{
\item \code{1} 0-4 persons
\item \code{2} 5-9
\item \code{3} 10-19
\item \code{4} 20-49
\item \code{5} 50 or more
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4126/vshow}{GSS Data Explorer page for \code{peocntct}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0-4 persons \tab 5-9 \tab 10-19 \tab 20-49 \tab 50 or more \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 247 \tab 354 \tab 378 \tab 336 \tab 199 \tab 2 \tab 2 \tab 4510 \cr
   2014 \tab 1274 \tab 278 \tab 358 \tab 286 \tab 216 \tab 115 \tab 10 \tab 1 \tab 2538 \cr
   Total \tab 4266 \tab 525 \tab 712 \tab 664 \tab 552 \tab 314 \tab 12 \tab 3 \tab 7048 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{macall}},
\code{\link{malive}},
\code{\link{matime}},
\code{\link{mavisit}},
\code{\link{palive}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
