% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{peopgovt}
\alias{peopgovt}
\title{People do more for environment than govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
peopgovt
}
\details{
Question C. And which of these two groups is making more effort to look after the environment:
}
\section{Values}{

\itemize{
\item \verb{[1]} people in general
\item \verb{[2]} or government
\item \verb{[3]} both equally
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4786/vshow}{GSS Data Explorer page for \code{peopgovt}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab people in general \tab or government \tab both equally \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 494 \tab 224 \tab 294 \tab 161 \tab 103 \tab 2817 \cr
   Total \tab 1541 \tab 494 \tab 224 \tab 294 \tab 161 \tab 103 \tab 2817 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
