% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{peopgrn}
\alias{peopgrn}
\title{How get people to protect environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
peopgrn
}
\details{
Question J. Which of these approaches do you think would be the best way of getting people and their families in America to protect the environment?
}
\section{Values}{

\itemize{
\item \code{1} heavy fines for people who damage the environment
\item \code{2} use the tax system to reward people who protect the environment
\item \code{3} more information and education for people about the advantages of protecting the environment
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4812/vshow}{GSS Data Explorer page for \code{peopgrn}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}
}
\concept{ISSP Environment}
\keyword{variable}
