% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_278.R
\name{peoplegn}
\alias{peoplegn}
\title{R tried to meet new people past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
peoplegn
}
\details{
Question In the last twelve months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/2002"have you:
D. Tried to meet new people for social purposes?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2535/vshow}{GSS Data Explorer page for \code{peoplegn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 1884 \tab 560 \tab 170 \tab 182 \tab 21 \tab - \tab - \tab 2817 \cr
   2002 \tab 1372 \tab 806 \tab 190 \tab 164 \tab 7 \tab 220 \tab 6 \tab 2765 \cr
   Total \tab 3256 \tab 1366 \tab 360 \tab 346 \tab 28 \tab 220 \tab 6 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinfgn}},
\code{\link{gamegn}},
\code{\link{investgn}},
\code{\link{newjobgn}},
\code{\link{polcamp}},
\code{\link{poldisgn}},
\code{\link{polfrel}},
\code{\link{polwww}},
\code{\link{travelgn}}
}
\concept{Information Society}
\keyword{variable}
