% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{peoptrbl}
\alias{peoptrbl}
\title{Assisting people in trouble is very important}
\source{
General Social Survey https://gss.norc.org
}
\description{
peoptrbl
}
\details{
Question Please tell me whether you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree with the following statements:
C. Personally assisting people in trouble is very important to me
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2892/vshow}{GSS Data Explorer page for \code{peoptrbl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 328 \tab 672 \tab 275 \tab 67 \tab 12 \tab 12 \tab 6 \tab 2765 \cr
   2004 \tab 1472 \tab 329 \tab 715 \tab 245 \tab 41 \tab 7 \tab 1 \tab 2 \tab 2812 \cr
   2012 \tab 666 \tab 277 \tab 748 \tab 218 \tab 47 \tab 3 \tab 7 \tab 8 \tab 1974 \cr
   2014 \tab 1264 \tab 326 \tab 706 \tab 179 \tab 41 \tab 11 \tab 5 \tab 6 \tab 2538 \cr
   Total \tab 4795 \tab 1260 \tab 2841 \tab 917 \tab 196 \tab 33 \tab 25 \tab 22 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{selffrst}},
\code{\link{selfless}}
}
\concept{Altruism}
\keyword{variable}
