% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_150.R
\name{perform}
\alias{perform}
\title{Performed music, dance, theater in lst yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
perform
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
J. Take part in a music, dance, or theatrical performance.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1414/vshow}{GSS Data Explorer page for \code{perform}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 154 \tab 1436 \tab 1 \tab 15 \tab 1606 \cr
   1998 \tab 1387 \tab 152 \tab 1280 \tab - \tab 13 \tab 2832 \cr
   2002 \tab 1393 \tab 141 \tab 1217 \tab 8 \tab 6 \tab 2765 \cr
   Total \tab 2780 \tab 447 \tab 3933 \tab 9 \tab 34 \tab 7203 \cr
}
}

\seealso{
Other Culture: 
\code{\link{camping}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{huntfish}},
\code{\link{plymusic}},
\code{\link{seemovie}},
\code{\link{usevcr}}
}
\concept{Culture}
\keyword{variable}
