% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{perfrmgn}
\alias{perfrmgn}
\title{R obtain information on performing arts past ye}
\source{
General Social Survey https://gss.norc.org
}
\description{
perfrmgn
}
\details{
Question In the last twelve months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/2002"have you:
I. Obtained information about a concert, play, or other performing-arts event?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2540/vshow}{GSS Data Explorer page for \code{perfrmgn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 1884 \tab 460 \tab 238 \tab 217 \tab 18 \tab - \tab - \tab 2817 \cr
   2002 \tab 1372 \tab 617 \tab 275 \tab 263 \tab 6 \tab 226 \tab 6 \tab 2765 \cr
   Total \tab 3256 \tab 1077 \tab 513 \tab 480 \tab 24 \tab 226 \tab 6 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{polcangn}},
\code{\link{polinfgn}}
}
\concept{Information Society}
\keyword{variable}
