% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{perfrt}
\alias{perfrt}
\title{Performance rating last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
perfrt
}
\details{
Question If you received a performance rating from your organization in 2021, which statement best describes the rating you received?
}
\section{Values}{

\itemize{
\item \code{1} below expectations
\item \code{2} meets expectations
\item \code{3} above expectations
\item \code{4} r didn't receive a performance rating
\item \code{5} none of the statements match the performance rating r received
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7853/vshow}{GSS Data Explorer page for \code{perfrt}}.
}

\keyword{variable}
