% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{perhrh}
\alias{perhrh}
\title{Hours per week use computer personal interest}
\source{
General Social Survey https://gss.norc.org
}
\description{
perhrh
}
\details{
Question About how many minutes or hours a week do you use your home computer for each of the following:
B. Personal interests, games, or enjoyment?
2. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2559/vshow}{GSS Data Explorer page for \code{perhrh}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 13 \tab 15 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2418 \tab 112 \tab 68 \tab 9 \tab 3 \tab 1 \tab 4 \tab 47 \tab 8 \tab 2 \tab 23 \tab 3 \tab 13 \tab 20 \tab 6 \tab 9 \tab 5 \tab 2 \tab 64 \tab 2817 \cr
   Total \tab 2418 \tab 112 \tab 68 \tab 9 \tab 3 \tab 1 \tab 4 \tab 47 \tab 8 \tab 2 \tab 23 \tab 3 \tab 13 \tab 20 \tab 6 \tab 9 \tab 5 \tab 2 \tab 64 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compwork}},
\code{\link{emhrh}},
\code{\link{emminh}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwhrh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
