% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{perhrw}
\alias{perhrw}
\title{Hours per week using work computer personal int}
\source{
General Social Survey https://gss.norc.org
}
\description{
perhrw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
D. Personal interests, games, or enjoyment?
2. Hours per week?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2574/vshow}{GSS Data Explorer page for \code{perhrw}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 8 \tab no answer \tab Total \cr
   2000 \tab 2486 \tab 224 \tab 13 \tab 3 \tab 9 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 74 \tab 2817 \cr
   Total \tab 2486 \tab 224 \tab 13 \tab 3 \tab 9 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 74 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{ecomminw}},
\code{\link{emhrw}},
\code{\link{emminw}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{perminw}},
\code{\link{wwwhrw}},
\code{\link{wwwminw}}
}
\concept{Information Society}
\keyword{variable}
