% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{perla}
\alias{perla}
\title{R skin tone via interviewer}
\source{
General Social Survey https://gss.norc.org
}
\description{
perla
}
\details{
Question USING THE SKIN TONE PALETTE, PLEASE SELECT THE OPTION THAT MOST MATCHES THE RESPONDENT’S FACIAL COLORING.
NOTE: Please see the 2024 Codebook for more information on this variable.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - lightest
\item \verb{[12]} 12 - darkest
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8287/vshow}{GSS Data Explorer page for \code{perla}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab not available in this year \tab 1 - lightest \tab 10 \tab 11 \tab 12 - darkest \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab iap \tab Total \cr
   2024 \tab - \tab 256 \tab 9 \tab 4 \tab 3 \tab 186 \tab 127 \tab 282 \tab 135 \tab 72 \tab 95 \tab 50 \tab 28 \tab 2062 \tab 3309 \cr
   Total \tab 0 \tab 256 \tab 9 \tab 4 \tab 3 \tab 186 \tab 127 \tab 282 \tab 135 \tab 72 \tab 95 \tab 50 \tab 28 \tab 2062 \tab 3309 \cr
}
}

\concept{Core}
\concept{Race}
\keyword{variable}
