% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{perla_self}
\alias{perla_self}
\title{R self-reported skin tone}
\source{
General Social Survey https://gss.norc.org
}
\description{
perla_self
}
\details{
Question Using the skin tone palette below, please select the option that most matches your skin tone.
NOTE: Please see the 2024 Codebook for more information on this variable. This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - lightest
\item \verb{[12]} 12 - darkest
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8288/vshow}{GSS Data Explorer page for \code{perla_self}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab not available in this year \tab 1 - lightest \tab 10 \tab 11 \tab 12 - darkest \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 186 \tab 13 \tab 5 \tab 1 \tab 182 \tab 89 \tab 310 \tab 98 \tab 50 \tab 47 \tab 15 \tab 20 \tab 42 \tab 2245 \tab 6 \tab 3309 \cr
   Total \tab 0 \tab 186 \tab 13 \tab 5 \tab 1 \tab 182 \tab 89 \tab 310 \tab 98 \tab 50 \tab 47 \tab 15 \tab 20 \tab 42 \tab 2245 \tab 6 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{happy_next}},
\code{\link{knwsexor}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}
}
\concept{GSS Next}
\concept{Race}
\keyword{variable}
