% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{permino}
\alias{permino}
\title{Minutes per week using other computer personal}
\source{
General Social Survey https://gss.norc.org
}
\description{
permino
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
B. Personal interests, games, or enjoyment?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2587/vshow}{GSS Data Explorer page for \code{permino}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 15 \tab 2 \tab 20 \tab 30 \tab 4 \tab 5 \tab 7 \tab no answer \tab Total \cr
   2000 \tab 2665 \tab 99 \tab 1 \tab 1 \tab 1 \tab 3 \tab 11 \tab 1 \tab 2 \tab 1 \tab 32 \tab 2817 \cr
   Total \tab 2665 \tab 99 \tab 1 \tab 1 \tab 1 \tab 3 \tab 11 \tab 1 \tab 2 \tab 1 \tab 32 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}}
}
\concept{Information Society}
\keyword{variable}
