% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{perminw}
\alias{perminw}
\title{Minutes per week using work computer personal i}
\source{
General Social Survey https://gss.norc.org
}
\description{
perminw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
D. Personal interests, games, or enjoyment?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2573/vshow}{GSS Data Explorer page for \code{perminw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 15 \tab 30 \tab 35 \tab 5 \tab no answer \tab Total \cr
   2000 \tab 2486 \tab 269 \tab 1 \tab 10 \tab 1 \tab 1 \tab 49 \tab 2817 \cr
   Total \tab 2486 \tab 269 \tab 1 \tab 10 \tab 1 \tab 1 \tab 49 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{ecomminw}},
\code{\link{emhrw}},
\code{\link{emminw}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{perhrw}},
\code{\link{wwwhrw}},
\code{\link{wwwminw}}
}
\concept{Information Society}
\keyword{variable}
