% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{permoral}
\alias{permoral}
\title{R agrees that morality a personal matter}
\source{
General Social Survey https://gss.norc.org
}
\description{
permoral
}
\details{
Question Please consider the following statements and tell me whether you agree strongly, agree somewhat, disagree somewhat or disagree strongly with each statement.
D. Morality is a personal matter and society should not force everyone to follow one standard.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1084/vshow}{GSS Data Explorer page for \code{permoral}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 469 \tab 575 \tab 259 \tab 113 \tab 57 \tab 8 \tab 1481 \cr
   2006 \tab 1518 \tab 1022 \tab 1158 \tab 519 \tab 204 \tab 79 \tab 10 \tab 4510 \cr
   2008 \tab - \tab 724 \tab 785 \tab 301 \tab 159 \tab 44 \tab 10 \tab 2023 \cr
   2010 \tab - \tab 707 \tab 796 \tab 332 \tab 146 \tab 52 \tab 11 \tab 2044 \cr
   Total \tab 1518 \tab 2922 \tab 3314 \tab 1411 \tab 622 \tab 232 \tab 39 \tab 10058 \cr
}
}

\seealso{
Other Religion: 
\code{\link{blkwhite}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{gochurch}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
