% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{perright}
\alias{perright}
\title{Conscience should decide right and wrong}
\source{
General Social Survey https://gss.norc.org
}
\description{
perright
}
\details{
Question Do you agree or disagree with the following statements. . .
C. Right and wrong should be a matter of personal conscience.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4683/vshow}{GSS Data Explorer page for \code{perright}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 180 \tab 393 \tab 255 \tab 211 \tab 188 \tab 60 \tab 72 \tab 1517 \cr
   Total \tab 158 \tab 180 \tab 393 \tab 255 \tab 211 \tab 188 \tab 60 \tab 72 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{socright}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
