% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_266.R
\name{person30}
\alias{person30}
\title{R use personal website in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
person30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
R. Personal home pages?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} 3-5 times
\item \code{4} more than 5 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2411/vshow}{GSS Data Explorer page for \code{person30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 458 \tab 127 \tab 30 \tab 53 \tab 74 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 742 \tab 148 \tab 38 \tab 60 \tab 94 \tab 2 \tab 2765 \cr
   2004 \tab 2122 \tab 525 \tab 78 \tab 27 \tab 45 \tab 12 \tab 3 \tab 2812 \cr
   Total \tab 5878 \tab 1725 \tab 353 \tab 95 \tab 158 \tab 180 \tab 5 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{invest12}},
\code{\link{porn30}},
\code{\link{sci30}},
\code{\link{work12}}
}
\concept{Information Society}
\keyword{variable}
