% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_339.R
\name{pessimst}
\alias{pessimst}
\title{I hardly ever expect things to go my way}
\source{
General Social Survey https://gss.norc.org
}
\description{
pessimst
}
\details{
Question Indicate your agreement with each of the following statements by selecting the number that comes closest to your answer:
G. I hardly ever expect things to go my way.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3042/vshow}{GSS Data Explorer page for \code{pessimst}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2004 \tab 432 \tab 98 \tab 416 \tab 1359 \tab 495 \tab 6 \tab 6 \tab 2812 \cr
   Total \tab 432 \tab 98 \tab 416 \tab 1359 \tab 495 \tab 6 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{afailure}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{satself}},
\code{\link{slfrspct}}

Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{satself}},
\code{\link{slfrspct}}
}
\concept{Genes and the Environment}
\concept{Work Organizations}
\keyword{variable}
