% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_317.R
\name{petb4}
\alias{petb4}
\title{Did r's family have a pet 5 years ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
petb4
}
\details{
Question Did your family have a pet 5 years ago?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7196/vshow}{GSS Data Explorer page for \code{petb4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 765 \tab 385 \tab 2 \tab 21 \tab 2348 \cr
   Total \tab 1175 \tab 765 \tab 385 \tab 2 \tab 21 \tab 2348 \cr
}
}

\keyword{variable}
