% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_320.R
\name{petb4ply}
\alias{petb4ply}
\title{How often r used to spend time each day playing with or training pet}
\source{
General Social Survey https://gss.norc.org
}
\description{
petb4ply
}
\details{
Question Thinking about your favorite pet, how often did you spend time each day playing with or exercising your pet? Would you say almost always, often, sometimes, or never?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7220/vshow}{GSS Data Explorer page for \code{petb4ply}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab never \tab no answer \tab Total \cr
   2018 \tab 2144 \tab 92 \tab 41 \tab 32 \tab 16 \tab 23 \tab 2348 \cr
   Total \tab 2144 \tab 92 \tab 41 \tab 32 \tab 16 \tab 23 \tab 2348 \cr
}
}

\keyword{variable}
