% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_319.R
\name{petfam}
\alias{petfam}
\title{How often r considers pet member of family}
\source{
General Social Survey https://gss.norc.org
}
\description{
petfam
}
\details{
Question How often do you consider your pet to be a member of your family?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7209/vshow}{GSS Data Explorer page for \code{petfam}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1649 \tab 522 \tab 80 \tab 38 \tab 31 \tab 1 \tab 27 \tab 2348 \cr
   Total \tab 1649 \tab 522 \tab 80 \tab 38 \tab 31 \tab 1 \tab 27 \tab 2348 \cr
}
}

\keyword{variable}
