% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{petition}
\alias{petition}
\title{Web make r sign petition}
\source{
General Social Survey https://gss.norc.org
}
\description{
petition
}
\details{
Question Have you ever taken any of the following actions because of something you learned about or discussed on the Web?
A. Signed a petition?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2442/vshow}{GSS Data Explorer page for \code{petition}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2453 \tab 59 \tab 221 \tab 84 \tab - \tab 2817 \cr
   2002 \tab 2411 \tab 91 \tab 217 \tab 45 \tab 1 \tab 2765 \cr
   Total \tab 4864 \tab 150 \tab 438 \tab 129 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
