% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{petsseq}
\alias{petsseq}
\title{Position of pet questions in questionnaire}
\source{
General Social Survey https://gss.norc.org
}
\description{
petsseq
}
\details{
Question Position of pet questions in questionnaire
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} beginning
\item \verb{[2]} middle
\item \verb{[3]} end
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8289/vshow}{GSS Data Explorer page for \code{petsseq}}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab beginning \tab middle \tab end \tab iap \tab Total \cr
   2024 \tab - \tab 385 \tab 348 \tab 332 \tab 2244 \tab 3309 \cr
   Total \tab 0 \tab 385 \tab 348 \tab 332 \tab 2244 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}}

Other GSS Next: 
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}}
}
\concept{GSS Next}
\keyword{variable}
