% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{phyeffrt}
\alias{phyeffrt}
\title{Rate physical effort}
\source{
General Social Survey https://gss.norc.org
}
\description{
phyeffrt
}
\details{
Question A. Please rate the overall physical effort at the job you normally Do.
}
\section{Values}{

\itemize{
\item \code{1} very hard
\item \code{2} hard
\item \code{3} somewhat hard
\item \code{4} fairly light
\item \code{5} very light
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2836/vshow}{GSS Data Explorer page for \code{phyeffrt}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{backpain}},
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{strredpg}},
\code{\link{usedup}}
}
\concept{Quality of Working Life}
\keyword{variable}
