% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{physacts}
\alias{physacts}
\title{To what extent r can accomplish everyday physical activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
physacts
}
\details{
Question To what extent are you able to carry out your everyday physical activities such as walking, climbing stairs, carrying groceries, or moving a chair?
}
\section{Values}{

\itemize{
\item \verb{[1]} completely
\item \verb{[2]} mostly
\item \verb{[3]} moderately
\item \verb{[4]} a little
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7168/vshow}{GSS Data Explorer page for \code{physacts}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab completely \tab mostly \tab moderately \tab a little \tab not at all \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 1499 \tab 407 \tab 276 \tab 129 \tab 18 \tab 2 \tab 17 \tab - \tab 2348 \cr
   2021 \tab 386 \tab 2190 \tab 667 \tab 507 \tab 219 \tab 40 \tab - \tab 1 \tab 22 \tab 4032 \cr
   Total \tab 386 \tab 3689 \tab 1074 \tab 783 \tab 348 \tab 58 \tab 2 \tab 18 \tab 22 \tab 6380 \cr
}
}

\keyword{variable}
