% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_376.R
\name{physcsci}
\alias{physcsci}
\title{How scientific: physics}
\source{
General Social Survey https://gss.norc.org
}
\description{
physcsci
}
\details{
Question Please look at Card B20. How scientific are each of the following fields? If you have not heard of a particular field, just say you haven’t heard of it
B. Physics. Is physics very scientific, pretty scientific, not too scientific, or not scientific at all?
}
\section{Values}{

\itemize{
\item \code{1} very scientific
\item \code{2} pretty scientific
\item \code{3} not too scientific
\item \code{4} not scientific at all
\item \code{5} haven't heard of it (volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3407/vshow}{GSS Data Explorer page for \code{physcsci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab haven't heard of it (volunteered) \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1262 \tab 394 \tab 59 \tab 15 \tab 50 \tab 80 \tab 4 \tab 4510 \cr
   2012 \tab 1477 \tab 331 \tab 102 \tab 21 \tab 8 \tab 12 \tab 17 \tab 6 \tab 1974 \cr
   Total \tab 4123 \tab 1593 \tab 496 \tab 80 \tab 23 \tab 62 \tab 97 \tab 10 \tab 6484 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{histsci}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{socsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
