% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{physwrk}
\alias{physwrk}
\title{How often must r do physical work}
\source{
General Social Survey https://gss.norc.org
}
\description{
physwrk
}
\details{
Question Now some more questions about your working conditions. one code for each item below to show how often it applies to your work. How often...
B. Do you have to do hard physical work?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} hardly ever
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4551/vshow}{GSS Data Explorer page for \code{physwrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 60 \tab 127 \tab 231 \tab 254 \tab 199 \tab 3 \tab 24 \tab 1537 \cr
   1998 \tab 1967 \tab 73 \tab 106 \tab 205 \tab 203 \tab 240 \tab 2 \tab 36 \tab 2832 \cr
   2006 \tab 3492 \tab 134 \tab 113 \tab 228 \tab 236 \tab 305 \tab - \tab 2 \tab 4510 \cr
   Total \tab 6098 \tab 267 \tab 346 \tab 664 \tab 693 \tab 744 \tab 5 \tab 62 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}},
\code{\link{xhaustn}}
}
\concept{Work Orientation}
\keyword{variable}
