% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{pilingup}
\alias{pilingup}
\title{How often r felt difficulty piling up in past 4 weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
pilingup
}
\details{
Question During the past four weeks, how often have you felt difficulties were piling up so high that you could not overcome them?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} rarely
\item \code{3} sometimes
\item \code{4} often
\item \code{5} very often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7273/vshow}{GSS Data Explorer page for \code{pilingup}}.
}

\keyword{variable}
