% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_287.R
\name{pinterst}
\alias{pinterst}
\title{Use pinterest}
\source{
General Social Survey https://gss.norc.org
}
\description{
pinterst
}
\details{
Question (Which of the following social networking or social media sites are you a member or regular user of?)
Pinterest.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6277/vshow}{GSS Data Explorer page for \code{pinterst}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab 1483 \tab 480 \tab 892 \tab 12 \tab - \tab 2867 \cr
   Total \tab 1483 \tab 480 \tab 892 \tab 12 \tab 0 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{clssmtes}},
\code{\link{facebook}},
\code{\link{flickr}},
\code{\link{googlesn}},
\code{\link{instagrm}},
\code{\link{intmbile}},
\code{\link{intrecnt}},
\code{\link{intuse}},
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{intwkenm}},
\code{\link{linkedin}},
\code{\link{snapchat}},
\code{\link{snsmfrst}},
\code{\link{snsmnew}},
\code{\link{snsmot2a}},
\code{\link{snsmot2b}},
\code{\link{snsmot2c}},
\code{\link{snsmoth1}},
\code{\link{tumblr}},
\code{\link{twitter}},
\code{\link{vine}},
\code{\link{whatsapp}}
}
\concept{Social Media Usage}
\keyword{variable}
