% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{plan1}
\alias{plan1}
\title{Who in household plans social activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
plan1
}
\details{
Question (In your household who does the following things...?) Plans and organizes social and family activities?
}
\section{Values}{

\itemize{
\item \verb{[1]} always me
\item \verb{[2]} usually me
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually my spouse/partner
\item \verb{[5]} always my spouse/partner
\item \verb{[6]} is done by a third person
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7775/vshow}{GSS Data Explorer page for \code{plan1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab always me \tab usually me \tab about equal or both together \tab usually my spouse/partner \tab always my spouse/partner \tab is done by a third person \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3024 \tab 63 \tab 114 \tab 205 \tab 91 \tab 31 \tab 2 \tab 11 \tab 1 \tab 2 \tab 3544 \cr
   Total \tab 3024 \tab 63 \tab 114 \tab 205 \tab 91 \tab 31 \tab 2 \tab 11 \tab 1 \tab 2 \tab 3544 \cr
}
}

\keyword{variable}
