% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{planetrp}
\alias{planetrp}
\title{Plane trip in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
planetrp
}
\details{
Question In the last twelve months, how many trips did you make by plane? Count outward and return journeys, including transfers, as one trip.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7497/vshow}{GSS Data Explorer page for \code{planetrp}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 15 \tab 16 \tab 180 \tab 2 \tab 20 \tab 200 \tab 24 \tab 25 \tab 28 \tab 3 \tab 30 \tab 4 \tab 40 \tab 5 \tab 6 \tab 60 \tab 7 \tab 8 \tab 80 \tab 9 \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 1275 \tab 205 \tab 7 \tab 4 \tab 1 \tab 1 \tab 1 \tab 149 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 50 \tab 1 \tab 61 \tab 1 \tab 7 \tab 24 \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 191 \tab 32 \tab 4032 \cr
   Total \tab 2003 \tab 1275 \tab 205 \tab 7 \tab 4 \tab 1 \tab 1 \tab 1 \tab 149 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 50 \tab 1 \tab 61 \tab 1 \tab 7 \tab 24 \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 191 \tab 32 \tab 4032 \cr
}
}

\keyword{variable}
