% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{planwrk}
\alias{planwrk}
\title{How much can r plan daily work}
\source{
General Social Survey https://gss.norc.org
}
\description{
planwrk
}
\details{
Question And which of the following statements about your work is most true?
}
\section{Values}{

\itemize{
\item \code{1} plans most
\item \code{2} plans part
\item \code{3} cant plan
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4557/vshow}{GSS Data Explorer page for \code{planwrk}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab plans most \tab plans part \tab cant plan \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 384 \tab 312 \tab 176 \tab 665 \tab - \tab 1537 \cr
   Total \tab 0 \tab 384 \tab 312 \tab 176 \tab 665 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
