% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_150.R
\name{plymusic}
\alias{plymusic}
\title{Played a musical instrument in lst yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
plymusic
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
N. Play a musical instrument like a piano, guitar, or violin.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1418/vshow}{GSS Data Explorer page for \code{plymusic}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 372 \tab 1218 \tab 1 \tab 15 \tab 1606 \cr
   1998 \tab 1387 \tab 337 \tab 1095 \tab - \tab 13 \tab 2832 \cr
   2002 \tab 1393 \tab 286 \tab 1072 \tab 8 \tab 6 \tab 2765 \cr
   Total \tab 2780 \tab 995 \tab 3385 \tab 9 \tab 34 \tab 7203 \cr
}
}

\seealso{
Other Culture: 
\code{\link{camping}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{huntfish}},
\code{\link{perform}},
\code{\link{seemovie}},
\code{\link{usevcr}}
}
\concept{Culture}
\keyword{variable}
