% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_624.R
\name{polabusey}
\alias{polabusey}
\title{Citizen said vulgar or obscene things y}
\source{
General Social Survey https://gss.norc.org
}
\description{
polabusey
}
\details{
Question (Would you approve of a policeman striking a citizen who...) Had said vulgar and obscene things to the police officer?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7448/vshow}{GSS Data Explorer page for \code{polabusey}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 2673 \tab 123 \tab 1228 \tab 1 \tab 7 \tab - \tab 4032 \cr
   2022 \tab 2364 \tab 119 \tab 1050 \tab 6 \tab 3 \tab 2 \tab 3544 \cr
   Total \tab 5037 \tab 242 \tab 2278 \tab 7 \tab 10 \tab 2 \tab 7576 \cr
}
}

\keyword{variable}
