% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{polactve}
\alias{polactve}
\title{Pol party encourage ppl to be active in politics in am}
\source{
General Social Survey https://gss.norc.org
}
\description{
polactve
}
\details{
Question Thinking now about politics in America, to what extent do you agree or disagree with the following statements?
A. Political parties encourage people to become active in politics
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4952/vshow}{GSS Data Explorer page for \code{polactve}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 178 \tab 738 \tab 317 \tab 194 \tab 23 \tab 15 \tab 7 \tab 2812 \cr
   2014 \tab 1274 \tab 92 \tab 508 \tab 346 \tab 198 \tab 40 \tab 74 \tab 6 \tab 2538 \cr
   Total \tab 2614 \tab 270 \tab 1246 \tab 663 \tab 392 \tab 63 \tab 89 \tab 13 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
