% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_378.R
\name{polaryr1}
\alias{polaryr1}
\title{Intl polar year 07-08:r would watch a tv program}
\source{
General Social Survey https://gss.norc.org
}
\description{
polaryr1
}
\details{
Question 2007-2008 has been designated the International Polar Year. During this year, many scientific and educational organizations will be making efforts to increase public awareness of the Earth's polar regions. These organizations hope that people will learn more about the people and animals that live near the poles, the changes taking place in the environment there, and new scientific discoveries about the Arctic and the Antarctic. Still looking at Card B22, do you think you would be very likely, somewhat likely, somewhat unlikely, or very unlikely to…
A. Watch a television program about the polar regions?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3424/vshow}{GSS Data Explorer page for \code{polaryr1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 642 \tab 727 \tab 287 \tab 187 \tab 19 \tab 2 \tab 4510 \cr
   2010 \tab 1338 \tab 244 \tab 290 \tab 88 \tab 70 \tab 5 \tab 9 \tab 2044 \cr
   Total \tab 3984 \tab 886 \tab 1017 \tab 375 \tab 257 \tab 24 \tab 11 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
