% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_378.R
\name{polaryr4}
\alias{polaryr4}
\title{Intl polar year 07-08:r would go to museum}
\source{
General Social Survey https://gss.norc.org
}
\description{
polaryr4
}
\details{
Question 2007-2008 has been designated the International Polar Year. During this year, many scientific and educational organizations will be making efforts to increase public awareness of the Earth’s polar regions. These organizations hope that people will learn more about the people and animals that live near the poles, the changes taking place in the environment there, and new scientific discoveries about the Arctic and the Antarctic. Still looking at Card B22, do you think you would be very likely, somewhat likely, somewhat unlikely, or very unlikely to…
D. Go to a museum to see an exhibit about the polar regions?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3427/vshow}{GSS Data Explorer page for \code{polaryr4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 309 \tab 561 \tab 496 \tab 478 \tab 18 \tab 2 \tab 4510 \cr
   2010 \tab 1338 \tab 135 \tab 235 \tab 162 \tab 164 \tab 3 \tab 7 \tab 2044 \cr
   Total \tab 3984 \tab 444 \tab 796 \tab 658 \tab 642 \tab 21 \tab 9 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr5}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
