% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_379.R
\name{polaryr8}
\alias{polaryr8}
\title{Intl polar year 07-08:r would attend a live lecture}
\source{
General Social Survey https://gss.norc.org
}
\description{
polaryr8
}
\details{
Question 2007-2008 has been designated the International Polar Year. During this year, many scientific and educational organizations will be making efforts to increase public awareness of the Earth’s polar regions. These organizations hope that people will learn more about the people and animals that live near the poles, the changes taking place in the environment there, and new scientific discoveries about the Arctic and the Antarctic. Still looking at Card B22, do you think you would be very likely, somewhat likely, somewhat unlikely, or very unlikely to…
H. Attend a live lecture by a scientist about the polar regions?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3431/vshow}{GSS Data Explorer page for \code{polaryr8}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 137 \tab 338 \tab 438 \tab 931 \tab 18 \tab 2 \tab 4510 \cr
   2010 \tab 1338 \tab 51 \tab 117 \tab 177 \tab 350 \tab 2 \tab 9 \tab 2044 \cr
   Total \tab 3984 \tab 188 \tab 455 \tab 615 \tab 1281 \tab 20 \tab 11 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{comorsci}},
\code{\link{extinct}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{nosun}},
\code{\link{polaryr6}},
\code{\link{polaryr7}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
