% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_624.R
\name{polattaky}
\alias{polattaky}
\title{Citizen attacking policeman with fists y}
\source{
General Social Survey https://gss.norc.org
}
\description{
polattaky
}
\details{
Question Would you approve if the citizen was attacking the police officer with his fists?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7449/vshow}{GSS Data Explorer page for \code{polattaky}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab don't know \tab Total \cr
   2021 \tab 2673 \tab 1031 \tab 323 \tab 5 \tab - \tab 4032 \cr
   2022 \tab 2364 \tab 925 \tab 238 \tab 2 \tab 15 \tab 3544 \cr
   Total \tab 5037 \tab 1956 \tab 561 \tab 7 \tab 15 \tab 7576 \cr
}
}

\keyword{variable}
