% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{polbstgw}
\alias{polbstgw}
\title{Best policy vs narrow interest: elected official}
\source{
General Social Survey https://gss.norc.org
}
\description{
polbstgw
}
\details{
Question Please look at Card B5. When making policy recommendations about global warming, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole versus what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
B. Elected Officials
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3363/vshow}{GSS Data Explorer page for \code{polbstgw}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 81 \tab 88 \tab 236 \tab 222 \tab 252 \tab 47 \tab 2 \tab 4510 \cr
   2010 \tab 1569 \tab 41 \tab 48 \tab 128 \tab 113 \tab 117 \tab 23 \tab 5 \tab 2044 \cr
   Total \tab 5151 \tab 122 \tab 136 \tab 364 \tab 335 \tab 369 \tab 70 \tab 7 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgw}},
\code{\link{bizinfgw}},
\code{\link{gasregs}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
