% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_373.R
\name{polbstsc}
\alias{polbstsc}
\title{Best policy on stem cell vs narrow interest: elected official}
\source{
General Social Survey https://gss.norc.org
}
\description{
polbstsc
}
\details{
Question Please look at Card B9. When making policy recommendations about stem cell research, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career
C. Elected officials
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3375/vshow}{GSS Data Explorer page for \code{polbstsc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 77 \tab 70 \tab 293 \tab 216 \tab 226 \tab 43 \tab 3 \tab 4510 \cr
   2010 \tab 1569 \tab 24 \tab 53 \tab 141 \tab 118 \tab 115 \tab 19 \tab 5 \tab 2044 \cr
   Total \tab 5151 \tab 101 \tab 123 \tab 434 \tab 334 \tab 341 \tab 62 \tab 8 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{medagrsc}},
\code{\link{medbstsc}},
\code{\link{medinfsc}},
\code{\link{polinfsc}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{scresrch}},
\code{\link{txbiz}},
\code{\link{txeco}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
