% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_268.R
\name{polcam12}
\alias{polcam12}
\title{Past 12 months r web to learn pol campaigns}
\source{
General Social Survey https://gss.norc.org
}
\description{
polcam12
}
\details{
Question In the past 12 months, have you used the Web to find out about or discuss?
G. Political campaigns?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2434/vshow}{GSS Data Explorer page for \code{polcam12}}.
}

\seealso{
Other Information Society: 
\code{\link{enviro12}},
\code{\link{fe12}},
\code{\link{foraff12}},
\code{\link{gun12}},
\code{\link{moral12}},
\code{\link{news12}},
\code{\link{polagree}},
\code{\link{racrel12}},
\code{\link{taxes12}}
}
\concept{Information Society}
\keyword{variable}
