% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{poleff10}
\alias{poleff10}
\title{The world is getting better}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff10
}
\details{
Question Please indicate whether you agree or disagree with each of the following statements.
J. Taking everything into account, the world is getting better.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4048/vshow}{GSS Data Explorer page for \code{poleff10}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 225 \tab 323 \tab 119 \tab 10 \tab 1534 \cr
   Total \tab 857 \tab 225 \tab 323 \tab 119 \tab 10 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{spenviro}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
