% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{poleff16}
\alias{poleff16}
\title{People elected to congress try to keep promises}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff16
}
\details{
Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE.
F. People we elect to Congress try to keep the promises they have made during the election
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1996 \cr
   A/C/- \tab 2012 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4100/vshow}{GSS Data Explorer page for \code{poleff16}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 45 \tab 277 \tab 271 \tab 502 \tab 178 \tab 33 \tab 26 \tab 2904 \cr
   2006 \tab 2992 \tab 33 \tab 296 \tab 301 \tab 536 \tab 333 \tab 19 \tab - \tab 4510 \cr
   2012 \tab 1312 \tab 10 \tab 140 \tab 119 \tab 253 \tab 117 \tab 20 \tab 3 \tab 1974 \cr
   2016 \tab 1477 \tab 36 \tab 247 \tab 303 \tab 540 \tab 240 \tab 12 \tab 12 \tab 2867 \cr
   Total \tab 7353 \tab 124 \tab 960 \tab 994 \tab 1831 \tab 868 \tab 84 \tab 41 \tab 12255 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{polgbeco}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
