% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{poleff18}
\alias{poleff18}
\title{Govt do not care much what ppl like r think}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff18
}
\details{
Question To what extent do you agree or disagree with the following statements?
A. I don’t think the government cares much what people like me think.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4939/vshow}{GSS Data Explorer page for \code{poleff18}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 227 \tab 423 \tab 273 \tab 470 \tab 73 \tab 2 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 308 \tab 413 \tab 213 \tab 238 \tab 64 \tab 25 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 535 \tab 836 \tab 486 \tab 708 \tab 137 \tab 27 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polgreed}},
\code{\link{polint1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
