% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{poleff19}
\alias{poleff19}
\title{R have a good understanding of pol issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff19
}
\details{
Question To what extent do you agree or disagree with the following statements?
B. I feel I have a pretty good understanding of the important political issues facing America.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4940/vshow}{GSS Data Explorer page for \code{poleff19}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 269 \tab 812 \tab 233 \tab 127 \tab 21 \tab 5 \tab 5 \tab 2812 \cr
   2014 \tab 1274 \tab 201 \tab 615 \tab 229 \tab 134 \tab 34 \tab 48 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 470 \tab 1427 \tab 462 \tab 261 \tab 55 \tab 53 \tab 8 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff18}},
\code{\link{poleff20}},
\code{\link{polgreed}},
\code{\link{polint1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
