% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_440.R
\name{poleff7}
\alias{poleff7}
\title{Govt is generally responsive to public}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff7
}
\details{
Question Please indicate whether you agree or disagree with each of the following statements.
G. The government is generally responsive to public opinion.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4045/vshow}{GSS Data Explorer page for \code{poleff7}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 302 \tab 294 \tab 70 \tab 11 \tab 1534 \cr
   Total \tab 857 \tab 302 \tab 294 \tab 70 \tab 11 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{mustret}},
\code{\link{poleff1}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff8}},
\code{\link{poleff9}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
