% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_440.R
\name{poleff9}
\alias{poleff9}
\title{People can control world affairs}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff9
}
\details{
Question Please indicate whether you agree or disagree with each of the following statements.
I. By taking an active part in political and social affairs the people can control world affairs.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4047/vshow}{GSS Data Explorer page for \code{poleff9}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{mustret}},
\code{\link{poleff1}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}}

Other Political: 
\code{\link{poleff1}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}}
}
\concept{ISSP Role of Gov't}
\concept{Political}
\keyword{variable}
