% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{polefy13}
\alias{polefy13}
\title{Understand issues facing country}
\source{
General Social Survey https://gss.norc.org
}
\description{
polefy13
}
\details{
Question How good is your understanding of the important political issues facing our country: Not at all, a little, somewhat, very, or extremely?
}
\section{Values}{

\itemize{
\item \verb{[1]} none
\item \verb{[2]} a little
\item \verb{[3]} some
\item \verb{[4]} quite a bit
\item \verb{[5]} a great deal
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3974/vshow}{GSS Data Explorer page for \code{polefy13}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab none \tab a little \tab some \tab quite a bit \tab a great deal \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1328 \tab 29 \tab 117 \tab 284 \tab 152 \tab 41 \tab 19 \tab 4 \tab 1974 \cr
   Total \tab 1328 \tab 29 \tab 117 \tab 284 \tab 152 \tab 41 \tab 19 \tab 4 \tab 1974 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{prespop}},
\code{\link{volacty2}},
\code{\link{volactyr}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
