% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{polefy15}
\alias{polefy15}
\title{How informed about politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
polefy15
}
\details{
Question Compared to most people, how informed are you about politics: not at all, a little, somewhat, very, or extremely?
}
\section{Values}{

\itemize{
\item \code{1} none
\item \code{2} a little
\item \code{3} some
\item \code{4} quite a bit
\item \code{5} a great deal
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3975/vshow}{GSS Data Explorer page for \code{polefy15}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab none \tab a little \tab some \tab quite a bit \tab a great deal \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1328 \tab 49 \tab 142 \tab 256 \tab 143 \tab 46 \tab 8 \tab 2 \tab 1974 \cr
   Total \tab 1328 \tab 49 \tab 142 \tab 256 \tab 143 \tab 46 \tab 8 \tab 2 \tab 1974 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{prespop}},
\code{\link{volacty2}},
\code{\link{volactyr}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
