% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{polefy17}
\alias{polefy17}
\title{How many govt admins can be trusted to do what's best}
\source{
General Social Survey https://gss.norc.org
}
\description{
polefy17
}
\details{
Question How many government administrators can be trusted to do what is best for the country: None, a few, some, most, or all?
}
\section{Values}{

\itemize{
\item \code{1} none
\item \code{2} a few
\item \code{3} some
\item \code{4} most
\item \code{5} all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3977/vshow}{GSS Data Explorer page for \code{polefy17}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy3}},
\code{\link{prespop}},
\code{\link{volacty2}},
\code{\link{volactyr}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
