% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{polgbeco}
\alias{polgbeco}
\title{What is your view on politics in us & world}
\source{
General Social Survey https://gss.norc.org
}
\description{
polgbeco
}
\details{
Question (HAND CARD C5)
Here are two options about what affects policies in America. Which of them comes closest to your view?
}
\section{Values}{

\itemize{
\item \verb{[1]} policies in america depend on what is happening in the world economy, rather than who is in government
\item \verb{[2]} policies in america depend more on who is in government, rather than what is happening in the world economy
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6310/vshow}{GSS Data Explorer page for \code{polgbeco}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab policies in america depend on what is happening in the world economy, rather than who is in government \tab policies in america depend more on who is in government, rather than what is happening in the world economy \tab can't choose \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 438 \tab 902 \tab 34 \tab 16 \tab 2867 \cr
   Total \tab 1477 \tab 438 \tab 902 \tab 34 \tab 16 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
