% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{polhitoy}
\alias{polhitoy}
\title{Ever approve of police striking citizen (form 2)}
\source{
General Social Survey https://gss.norc.org
}
\description{
polhitoy
}
\details{
Question Are there any situations you can imagine in which you would approve of a policeman striking an adult male citizen or are there no situations you can imagine in which you would approve of this?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5572/vshow}{GSS Data Explorer page for \code{polhitoy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1338 \tab 480 \tab 176 \tab 28 \tab 1 \tab 2023 \cr
   Total \tab 1338 \tab 480 \tab 176 \tab 28 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Form Y: 
\code{\link{letdie1y}},
\code{\link{popespky}}
}
\concept{Form Y}
\keyword{variable}
