% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{polinfgn}
\alias{polinfgn}
\title{R sought political information past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinfgn
}
\details{
Question In the last twelve months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/2002–have you:
H. Looked for information about political or social policy issues, current affairs, or political campaigns?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2539/vshow}{GSS Data Explorer page for \code{polinfgn}}.
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}}

Other Leisure: 
\code{\link{artgn}},
\code{\link{perfrmgn}},
\code{\link{polcangn}}
}
\concept{Information Society}
\concept{Leisure}
\keyword{variable}
