% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{polinfgw}
\alias{polinfgw}
\title{Influence on global warming policy: elected official}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinfgw
}
\details{
Question Please look at Card B4. How much influence should each of the following groups have in deciding what  to do about global warming?
B. Elected officials. Would you say a great deal of influence, a fair amount, a little influence, or none at all?
}
\section{Values}{

\itemize{
\item \code{1} a great deal of influence
\item \code{2} a fair amount
\item \code{3} a little influence
\item \code{4} none at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3360/vshow}{GSS Data Explorer page for \code{polinfgw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal of influence \tab a fair amount \tab a little influence \tab none at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 150 \tab 300 \tab 305 \tab 132 \tab 39 \tab 2 \tab 4510 \cr
   2010 \tab 1569 \tab 58 \tab 162 \tab 173 \tab 63 \tab 15 \tab 4 \tab 2044 \cr
   Total \tab 5151 \tab 208 \tab 462 \tab 478 \tab 195 \tab 54 \tab 6 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgw}},
\code{\link{bizinfgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
