% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{polinfnk}
\alias{polinfnk}
\title{Elected officials how much influence}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinfnk
}
\details{
Question Some say that the government should expand the use of nuclear power to generate electricity in the U.S. Others oppose any expansion in the use of nuclear power in this country. Please look at Card E15. How much influence should each of the following groups have in deciding whether to expand the use of nuclear power?
Elected officials. (Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal of influence
\item \verb{[2]} a fair amount
\item \verb{[3]} a little influence
\item \verb{[4]} none at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3493/vshow}{GSS Data Explorer page for \code{polinfnk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal of influence \tab a fair amount \tab a little influence \tab none at all \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1578 \tab 45 \tab 179 \tab 162 \tab 52 \tab 21 \tab 7 \tab 2044 \cr
   Total \tab 1578 \tab 45 \tab 179 \tab 162 \tab 52 \tab 21 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstnk}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{nukeelec}},
\code{\link{polbstnk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
