% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_277.R
\name{polinfo}
\alias{polinfo}
\title{R sought political information past 2 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinfo
}
\details{
Question In the past two years, that is, between (INSERT CURRENT MONTH) 1998 and (INSERT CURRENT MONTH) 2000"have you looked for information about the views or background of a candidate for political office?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2524/vshow}{GSS Data Explorer page for \code{polinfo}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3 or more times \tab no answer \tab Total \cr
   2000 \tab 1884 \tab 674 \tab 111 \tab 137 \tab 11 \tab 2817 \cr
   Total \tab 1884 \tab 674 \tab 111 \tab 137 \tab 11 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthmag2}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polpapr}},
\code{\link{poltv}}
}
\concept{Information Society}
\keyword{variable}
