% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_373.R
\name{polinfsc}
\alias{polinfsc}
\title{Influence on funding for stem cell rsch: elected officials}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinfsc
}
\details{
Question Please look at Card B8. How much influence should each of the following groups have in deciding about government funding for stem cell research?
C. Elected Officials. (Would you say a great deal of influence, a fair amount, a little influence, or none at all?)
}
\section{Values}{

\itemize{
\item \code{1} a great deal of influence
\item \code{2} a fair amount
\item \code{3} a little influence
\item \code{4} none at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3372/vshow}{GSS Data Explorer page for \code{polinfsc}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{medagrsc}},
\code{\link{medbstsc}},
\code{\link{medinfsc}},
\code{\link{polbstsc}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{scresrch}},
\code{\link{txbiz}},
\code{\link{txeco}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
