% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_374.R
\name{polinftx}
\alias{polinftx}
\title{Influence on reducing fed inc taxby elected officials}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinftx
}
\details{
Question Please look at Card B12. How much influence should each of the following groups have in deciding whether to reduce federal income taxes?
C. Elected officials. (Would you say a great deal of influence, a fair amount, a little influence, or none at all?)
}
\section{Values}{

\itemize{
\item \code{1} a great deal of influence
\item \code{2} a fair amount
\item \code{3} a little influence
\item \code{4} none at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3383/vshow}{GSS Data Explorer page for \code{polinftx}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal of influence \tab a fair amount \tab a little influence \tab none at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3574 \tab 203 \tab 350 \tab 236 \tab 101 \tab 46 \tab - \tab 4510 \cr
   2010 \tab 1578 \tab 101 \tab 186 \tab 112 \tab 41 \tab 20 \tab 6 \tab 2044 \cr
   Total \tab 5152 \tab 304 \tab 536 \tab 348 \tab 142 \tab 66 \tab 6 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbsttx}},
\code{\link{bizinftx}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{polbsttx}},
\code{\link{txpol}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
