% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{polint1}
\alias{polint1}
\title{How interested in politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
polint1
}
\details{
Question How interested would you say you personally are in politics?
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} fairly interested
\item \verb{[3]} not very interested
\item \verb{[4]} not at all interested
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4944/vshow}{GSS Data Explorer page for \code{polint1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very interested \tab fairly interested \tab not very interested \tab not at all interested \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 304 \tab 717 \tab 299 \tab 145 \tab 2 \tab 5 \tab 2812 \cr
   2014 \tab 1274 \tab 186 \tab 525 \tab 354 \tab 174 \tab 23 \tab 2 \tab 2538 \cr
   Total \tab 2614 \tab 490 \tab 1242 \tab 653 \tab 319 \tab 25 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polgreed}}
}
\concept{ISSP Citizenship}
\keyword{variable}
