% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{polintrst}
\alias{polintrst}
\title{Interested in campaigns this year}
\source{
General Social Survey https://gss.norc.org
}
\description{
polintrst
}
\details{
Question Some people don't pay much attention to political campaigns. How about you? Would you say that you have been very much interested, somewhat interested or not much interested in the political campaigns so far this year?
}
\section{Values}{

\itemize{
\item \verb{[1]} very much interested
\item \verb{[2]} somewhat interested
\item \verb{[3]} not much interested
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8158/vshow}{GSS Data Explorer page for \code{polintrst}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab very much interested \tab somewhat interested \tab not much interested \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 861 \tab 1258 \tab 1127 \tab 14 \tab 16 \tab 12 \tab 21 \tab 3309 \cr
   Total \tab 0 \tab 861 \tab 1258 \tab 1127 \tab 14 \tab 16 \tab 12 \tab 21 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{whovote24a}}
}
\concept{ANES}
\keyword{variable}
