% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_75.R
\name{polmurdr}
\alias{polmurdr}
\title{Citizen questioned as murder suspect}
\source{
General Social Survey https://gss.norc.org
}
\description{
polmurdr
}
\details{
Question Are there any situations you can imagine in which you would approve of a policeman striking an adult male citizen?
B. Was being questioned as a suspect in a murder case?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/672/vshow}{GSS Data Explorer page for \code{polmurdr}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1973 \tab 380 \tab 94 \tab 1013 \tab 12 \tab 5 \tab - \tab 1504 \cr
   1975 \tab 343 \tab 86 \tab 1030 \tab 24 \tab 7 \tab - \tab 1490 \cr
   1976 \tab - \tab 114 \tab 1351 \tab 29 \tab 5 \tab - \tab 1499 \cr
   1978 \tab - \tab 123 \tab 1357 \tab 50 \tab 2 \tab - \tab 1532 \cr
   1980 \tab - \tab 122 \tab 1308 \tab 31 \tab 7 \tab - \tab 1468 \cr
   1983 \tab - \tab 145 \tab 1422 \tab 30 \tab 2 \tab - \tab 1599 \cr
   1984 \tab - \tab 129 \tab 1301 \tab 28 \tab 15 \tab - \tab 1473 \cr
   1986 \tab - \tab 127 \tab 1319 \tab 19 \tab 5 \tab - \tab 1470 \cr
   1987 \tab - \tab 170 \tab 1588 \tab 53 \tab 8 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 84 \tab 887 \tab 21 \tab 5 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 81 \tab 924 \tab 25 \tab 5 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 101 \tab 771 \tab 26 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 66 \tab 925 \tab 19 \tab 7 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 74 \tab 951 \tab 28 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 136 \tab 1829 \tab 34 \tab 12 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 92 \tab 1796 \tab 31 \tab 6 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 107 \tab 1766 \tab 33 \tab 5 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 108 \tab 1747 \tab 29 \tab 12 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 81 \tab 806 \tab 25 \tab - \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 87 \tab 756 \tab 31 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 235 \tab 1681 \tab 73 \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 189 \tab 1138 \tab 38 \tab - \tab - \tab 2023 \cr
   2010 \tab 667 \tab 190 \tab 1141 \tab 43 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 192 \tab 1106 \tab 38 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 241 \tab 1412 \tab 34 \tab - \tab - \tab 2538 \cr
   2016 \tab 911 \tab 319 \tab 1585 \tab 45 \tab 7 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 253 \tab 1260 \tab 48 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 560 \tab 2074 \tab 2 \tab - \tab 36 \tab 4032 \cr
   2022 \tab 1173 \tab 454 \tab 1878 \tab 23 \tab 4 \tab 12 \tab 3544 \cr
   Total \tab 20384 \tab 4760 \tab 38122 \tab 922 \tab 133 \tab 48 \tab 64369 \cr
}
}

\seealso{
Other Core: 
\code{\link{gunnum}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{polabuse}},
\code{\link{polhitok}}
}
\concept{Core}
\keyword{variable}
