% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{polneutl}
\alias{polneutl}
\title{Visit politically neutral sites}
\source{
General Social Survey https://gss.norc.org
}
\description{
polneutl
}
\details{
Question Is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that you:
B. Visit sites that provide information without promoting a particular political point of view?
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2440/vshow}{GSS Data Explorer page for \code{polneutl}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not true \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 85 \tab 108 \tab 79 \tab 7 \tab 85 \tab 2817 \cr
   2002 \tab 2411 \tab 96 \tab 109 \tab 102 \tab 2 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 181 \tab 217 \tab 181 \tab 9 \tab 130 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{petition}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polnew}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
