% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{polnew}
\alias{polnew}
\title{R learn something new on the web}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnew
}
\details{
Question Now I would like you to think about the last time you used the Web to get information or exchange views about a political or social-policy issue, current affairs, or a political campaign.
A. Thinking back to this occasion, did you learn anything new about the issue or campaign?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{8} cant recall
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2445/vshow}{GSS Data Explorer page for \code{polnew}}.
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{petition}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polneutl}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
