% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{polnews}
\alias{polnews}
\title{How often use media to get political news}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnews
}
\details{
Question How often do you use the media, including television, newspapers, radio and the internet, to get political news or information?
}
\section{Values}{

\itemize{
\item \code{1} several times a day
\item \code{2} once a day
\item \code{3} 5-6 days a week
\item \code{4} 3-4 days a week
\item \code{5} 1-2 days a week
\item \code{6} less than 1 day a week
\item \code{7} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4966/vshow}{GSS Data Explorer page for \code{polnews}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}}

Other Political: 
\code{\link{corruptn}},
\code{\link{fixmistk}},
\code{\link{polinter}}
}
\concept{ISSP Citizenship}
\concept{Political}
\keyword{variable}
