% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{polnewsfrom}
\alias{polnewsfrom}
\title{Where r gets political news}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnewsfrom
}
\details{
Question Where do you get most of your news about political matters? Please select the most important source
}
\section{Values}{

\itemize{
\item \verb{[1]} online social media (for example facebook, instagram, tiktok, x (formerly twitter), youtube, snapchat)
\item \verb{[2]} television (including tv apps or access via internet)
\item \verb{[3]} newspapers (in print or access via internet)
\item \verb{[4]} radio and podcasts (including internet radio or access via internet
\item \verb{[5]} websites (including exclusively online news sites, news aggregators, portals, or search engines)
\item \verb{[6]} other
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8159/vshow}{GSS Data Explorer page for \code{polnewsfrom}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab online social media (for example facebook, instagram, tiktok, x (formerly twitter), youtube, snapchat) \tab television (including tv apps or access via internet) \tab newspapers (in print or access via internet) \tab radio and podcasts (including internet radio or access via internet \tab websites (including exclusively online news sites, news aggregators, portals, or search engines) \tab other \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 398 \tab 562 \tab 130 \tab 103 \tab 225 \tab 66 \tab 136 \tab 1670 \tab 6 \tab 13 \tab 3309 \cr
   Total \tab 0 \tab 398 \tab 562 \tab 130 \tab 103 \tab 225 \tab 66 \tab 136 \tab 1670 \tab 6 \tab 13 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
