% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{polnuke}
\alias{polnuke}
\title{How well elected official know risk by nuclear power}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnuke
}
\details{
Question Here’s another public issue: nuclear power. Nuclear power uses the energy of the atom to generate electricity. some people say that nuclear power is a safe option to meet the energy needs of the United States without burning fossil fuels such as coal, oil, or natural gas. Others are concerned about potential risks in the operation of nuclear plants and about the disposal of nuclear waste. Please look at Card E13. (On a scale of 1 to 5, where 1 means "Very Well" and 5 means  Not At All), how well do the following groups understand the likely effects of using nuclear power to generate electricity?
A. Elected officials (How well elected officials know risk by nuclear power?)
}
\section{Values}{

\itemize{
\item \code{1} very well
\item \code{5} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3488/vshow}{GSS Data Explorer page for \code{polnuke}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}

Other Environment: 
\code{\link{morempg}}
}
\concept{Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
