% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{polopts}
\alias{polopts}
\title{How important: ppl given chance to participate in decsns}
\source{
General Social Survey https://gss.norc.org
}
\description{
polopts
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
E. That people be given more opportunities to participate in public decision-making
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4937/vshow}{GSS Data Explorer page for \code{polopts}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2004 \tab 1340 \tab 15 \tab 9 \tab 23 \tab 89 \tab 243 \tab 254 \tab 822 \tab 12 \tab 5 \tab - \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 13 \tab 10 \tab 18 \tab 92 \tab 196 \tab 258 \tab 639 \tab 36 \tab 2 \tab - \tab - \tab 2538 \cr
   2024 \tab 1663 \tab 9 \tab 9 \tab 26 \tab 99 \tab 227 \tab 321 \tab 855 \tab 75 \tab 4 \tab - \tab 21 \tab 3309 \cr
   Total \tab 4277 \tab 37 \tab 28 \tab 67 \tab 280 \tab 666 \tab 833 \tab 2316 \tab 123 \tab 11 \tab 0 \tab 21 \tab 8659 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{buypol}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{citviews}},
\code{\link{cntctgov}},
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{eqtreat}},
\code{\link{fixmistk}},
\code{\link{govdook}},
\code{\link{govngos}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{gvtrghts}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{hlthall}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{leftrght}},
\code{\link{leftrght1}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ntcitvte}},
\code{\link{obeylaws}},
\code{\link{oppsegov}},
\code{\link{othreasn}},
\code{\link{owndecns}},
\code{\link{paytaxes}},
\code{\link{polactve}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polnews}},
\code{\link{poltalk}},
\code{\link{powerun}},
\code{\link{racmeet}},
\code{\link{refrndms}},
\code{\link{relmeet}},
\code{\link{revmeet}},
\code{\link{rghtsmin}},
\code{\link{servepeo}},
\code{\link{servpub}},
\code{\link{signdpet}},
\code{\link{solok}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usmca1}},
\code{\link{usmca2}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Democracy: 
\code{\link{citviews}},
\code{\link{crimlose}},
\code{\link{eqtreat}},
\code{\link{oppsegov}},
\code{\link{rghtsmin}},
\code{\link{solok}}
}
\concept{Democracy}
\concept{ISSP Citizenship}
\keyword{variable}
