% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_277.R
\name{polpapr}
\alias{polpapr}
\title{R sought political information daily newspaper}
\source{
General Social Survey https://gss.norc.org
}
\description{
polpapr
}
\details{
Question If YES, please tell me if you looked for such candidate information from:
A. Articles in a daily newspaper?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2525/vshow}{GSS Data Explorer page for \code{polpapr}}.
}

\seealso{
Other Information Society: 
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthmag2}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{poltv}}
}
\concept{Information Society}
\keyword{variable}
