% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{polsfair}
\alias{polsfair}
\title{Opinion of fairness of public official s treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
polsfair
}
\details{
Question In your opinion, how often do public officials deal fairly with people like you?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} often
\item \verb{[3]} occasionally
\item \verb{[4]} seldom
\item \verb{[5]} almost never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4121/vshow}{GSS Data Explorer page for \code{polsfair}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost always \tab often \tab occasionally \tab seldom \tab almost never \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 162 \tab 326 \tab 510 \tab 293 \tab 183 \tab 41 \tab 3 \tab 4510 \cr
   Total \tab 2992 \tab 162 \tab 326 \tab 510 \tab 293 \tab 183 \tab 41 \tab 3 \tab 4510 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{wotrial}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
