% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{polsgod}
\alias{polsgod}
\title{Atheists shouldn't hold public office}
\source{
General Social Survey https://gss.norc.org
}
\description{
polsgod
}
\details{
Question How much do you agree or disagree with each of the following…
A. Politicians who do not believe in God are unfit for public office.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4645/vshow}{GSS Data Explorer page for \code{polsgod}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 188 \tab 197 \tab 354 \tab 400 \tab 138 \tab 57 \tab 25 \tab 1517 \cr
   Total \tab 158 \tab 188 \tab 197 \tab 354 \tab 400 \tab 138 \tab 57 \tab 25 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
