% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{poltalk}
\alias{poltalk}
\title{Politicians talk too much and don't take action}
\source{
General Social Survey https://gss.norc.org
}
\description{
poltalk
}
\details{
Question To what extent do you agree or disagree with the following statements?… Politicians talk too much and take too little action.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8060/vshow}{GSS Data Explorer page for \code{poltalk}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 589 \tab 740 \tab 167 \tab 37 \tab 6 \tab 80 \tab 1667 \tab 4 \tab 19 \tab 3309 \cr
   Total \tab 0 \tab 589 \tab 740 \tab 167 \tab 37 \tab 6 \tab 80 \tab 1667 \tab 4 \tab 19 \tab 3309 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{buypol}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{citviews}},
\code{\link{cntctgov}},
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{eqtreat}},
\code{\link{fixmistk}},
\code{\link{govdook}},
\code{\link{govngos}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{gvtrghts}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{hlthall}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{leftrght}},
\code{\link{leftrght1}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ntcitvte}},
\code{\link{obeylaws}},
\code{\link{oppsegov}},
\code{\link{othreasn}},
\code{\link{owndecns}},
\code{\link{paytaxes}},
\code{\link{polactve}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polnews}},
\code{\link{polopts}},
\code{\link{powerun}},
\code{\link{racmeet}},
\code{\link{refrndms}},
\code{\link{relmeet}},
\code{\link{revmeet}},
\code{\link{rghtsmin}},
\code{\link{servepeo}},
\code{\link{servpub}},
\code{\link{signdpet}},
\code{\link{solok}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usmca1}},
\code{\link{usmca2}},
\code{\link{voteelec}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
