% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{poltrtblk}
\alias{poltrtblk}
\title{Police treats whites better than black people}
\source{
General Social Survey https://gss.norc.org
}
\description{
poltrtblk
}
\details{
Question Do the police treat Whites better than Black people, treat both the same, or treat Black people better than Whites?
}
\section{Values}{

\itemize{
\item \code{1} treat whites much better than blacks
\item \code{2} treat whites moderately better than blacks
\item \code{3} treat whites a little better than blacks
\item \code{4} treat both the same
\item \code{5} treat blacks a little better than whites
\item \code{6} treat blacks moderately better than whites
\item \code{7} treat blacks much better than whites
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7562/vshow}{GSS Data Explorer page for \code{poltrtblk}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab treat whites much better than blacks \tab treat whites moderately better than blacks \tab treat whites a little better than blacks \tab treat both the same \tab treat blacks a little better than whites \tab treat blacks moderately better than whites \tab treat blacks much better than whites \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1315 \tab 914 \tab 452 \tab 167 \tab 765 \tab 19 \tab 5 \tab 6 \tab 10 \tab 365 \tab 14 \tab 4032 \cr
   Total \tab 1315 \tab 914 \tab 452 \tab 167 \tab 765 \tab 19 \tab 5 \tab 6 \tab 10 \tab 365 \tab 14 \tab 4032 \cr
}
}

\keyword{variable}
