% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{poltrthsp}
\alias{poltrthsp}
\title{Police treats whites better than latinos or latinos better than whites}
\source{
General Social Survey https://gss.norc.org
}
\description{
poltrthsp
}
\details{
Question In general, do the police treat Whites better than Latinos, treat both the same, or treak Latinos better than Whites?
}
\section{Values}{

\itemize{
\item \code{1} treat whites much better than latinos
\item \code{2} treat whites moderately better than latinos
\item \code{3} treat whites a little better than latinos
\item \code{4} treat both the same
\item \code{5} treat latinos a little better than whites
\item \code{6} treat latinos moderately better than whites
\item \code{7} treat latinos much better than whites
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7568/vshow}{GSS Data Explorer page for \code{poltrthsp}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab treat whites much better than latinos \tab treat whites moderately better than latinos \tab treat whites a little better than latinos \tab treat both the same \tab treat latinos a little better than whites \tab treat latinos moderately better than whites \tab treat latinos much better than whites \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1315 \tab 761 \tab 509 \tab 174 \tab 856 \tab 3 \tab 6 \tab 6 \tab 12 \tab 373 \tab 17 \tab 4032 \cr
   Total \tab 1315 \tab 761 \tab 509 \tab 174 \tab 856 \tab 3 \tab 6 \tab 6 \tab 12 \tab 373 \tab 17 \tab 4032 \cr
}
}

\keyword{variable}
