% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{polviewx}
\alias{polviewx}
\title{Liberal or conservative-version x}
\source{
General Social Survey https://gss.norc.org
}
\description{
polviewx
}
\details{
Question C. In political matters, people talk of "the left" and "the right."  How would you place your views on this scale?
(SHOW CARD S. Do NOT PROMPT. IF RESPONDENT HESITATES, ASK R TO TRY AGAIN. MARK ANSWER BELOW.)
}
\section{Values}{

\itemize{
\item \verb{[1]} left
\item \verb{[10]} right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/180/vshow}{GSS Data Explorer page for \code{polviewx}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab left \tab right \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1983 \tab 804 \tab 16 \tab 19 \tab 17 \tab 46 \tab 77 \tab 311 \tab 102 \tab 82 \tab 46 \tab 14 \tab 55 \tab 10 \tab 1599 \cr
   Total \tab 804 \tab 16 \tab 19 \tab 17 \tab 46 \tab 77 \tab 311 \tab 102 \tab 82 \tab 46 \tab 14 \tab 55 \tab 10 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewy}},
\code{\link{pres12}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
