% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{polviewy}
\alias{polviewy}
\title{Liberal or conservative-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
polviewy
}
\details{
Question B. We hear a lot of talk these days about liberals and conservatives.
I'm going to show you a seven-point scale on which the political views that people might hold are arranged from extremely liberal--point 1--to extremely conservative--point 7. Where would you place yourself on this scale, or haven't you thought much about this?
}
\section{Values}{

\itemize{
\item \verb{[0]} havent thght abt it
\item \verb{[1]} extremely liberal
\item \verb{[2]} liberal
\item \verb{[3]} slightly liberal
\item \verb{[4]} moderate
\item \verb{[5]} slghtly conservative
\item \verb{[6]} conservative
\item \verb{[7]} extrmly conservative
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/179/vshow}{GSS Data Explorer page for \code{polviewy}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab havent thght abt it \tab extremely liberal \tab liberal \tab slightly liberal \tab moderate \tab slghtly conservative \tab conservative \tab extrmly conservative \tab don't know \tab no answer \tab Total \cr
   1978 \tab 761 \tab 124 \tab 12 \tab 66 \tab 109 \tab 239 \tab 116 \tab 74 \tab 11 \tab 17 \tab 3 \tab 1532 \cr
   Total \tab 761 \tab 124 \tab 12 \tab 66 \tab 109 \tab 239 \tab 116 \tab 74 \tab 11 \tab 17 \tab 3 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{pres12}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
