% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_278.R
\name{polwww}
\alias{polwww}
\title{R sought political information www}
\source{
General Social Survey https://gss.norc.org
}
\description{
polwww
}
\details{
Question If YES, please tell me if you looked for such candidate information from:
G. The Internet or World Wide Web?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2531/vshow}{GSS Data Explorer page for \code{polwww}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3 or more times \tab no answer \tab Total \cr
   2000 \tab 2558 \tab 150 \tab 46 \tab 49 \tab 14 \tab 2817 \cr
   Total \tab 2558 \tab 150 \tab 46 \tab 49 \tab 14 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinfgn}},
\code{\link{gamegn}},
\code{\link{investgn}},
\code{\link{newjobgn}},
\code{\link{peoplegn}},
\code{\link{polcamp}},
\code{\link{poldisgn}},
\code{\link{polfrel}},
\code{\link{travelgn}}
}
\concept{Information Society}
\keyword{variable}
