% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{popespks}
\alias{popespks}
\title{Pope is infallible on matter's of faith or morals}
\source{
General Social Survey https://gss.norc.org
}
\description{
popespks
}
\details{
Question We are interested in what American Catholics think about religious matters. please select the answer from this card that comes closest to your own personal opinion about the following statement: Under certain conditions, the pope is infallible when he speaks on matters of faith and morals.
}
\section{Values}{

\itemize{
\item \code{1} certainly true
\item \code{2} probably true
\item \code{3} i am uncertain whether this is true or false
\item \code{4} probably false
\item \code{5} certainly false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/340/vshow}{GSS Data Explorer page for \code{popespks}}.
}

\seealso{
Other Core: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{rdeemlib}}
}
\concept{Core}
\keyword{variable}
