% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{popespky}
\alias{popespky}
\title{Pope is infallible on matter's of faith or morals (form 2)}
\source{
General Social Survey https://gss.norc.org
}
\description{
popespky
}
\details{
Question Under certain conditions, the pope is infallible when he speaks on matters of faith and morals.
}
\section{Values}{

\itemize{
\item \verb{[1]} certainly false
\item \verb{[2]} probably false
\item \verb{[3]} i am uncertain whether this is false or true
\item \verb{[4]} probably true
\item \verb{[5]} certainly true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5571/vshow}{GSS Data Explorer page for \code{popespky}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab certainly false \tab probably false \tab i am uncertain whether this is false or true \tab probably true \tab certainly true \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1779 \tab 11 \tab 20 \tab 93 \tab 65 \tab 43 \tab 7 \tab 5 \tab 2023 \cr
   Total \tab 1779 \tab 11 \tab 20 \tab 93 \tab 65 \tab 43 \tab 7 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Form Y: 
\code{\link{letdie1y}},
\code{\link{polhitoy}}
}
\concept{Form Y}
\keyword{variable}
