% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{popgrwth}
\alias{popgrwth}
\title{Earth cannot support population growth}
\source{
General Social Survey https://gss.norc.org
}
\description{
popgrwth
}
\details{
Question And please circle one number for each of these statements to show how much you agree or disagree with it. The earth cannot continue to support population growth at its present rate.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4778/vshow}{GSS Data Explorer page for \code{popgrwth}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 196 \tab 441 \tab 279 \tab 197 \tab 37 \tab 55 \tab 71 \tab 2817 \cr
   2010 \tab 614 \tab 147 \tab 557 \tab 280 \tab 307 \tab 36 \tab 97 \tab 6 \tab 2044 \cr
   Total \tab 2155 \tab 343 \tab 998 \tab 559 \tab 504 \tab 73 \tab 152 \tab 77 \tab 4861 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}}
}
\concept{ISSP Environment}
\keyword{variable}
