% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{popmusic}
\alias{popmusic}
\title{Went to a live performance of popular music}
\source{
General Social Survey https://gss.norc.org
}
\description{
popmusic
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
B. Go to a live performance of popular music like rock, country, or rap, not including school performances.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2130/vshow}{GSS Data Explorer page for \code{popmusic}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab not available in this year \tab Total \cr
   1998 \tab 1387 \tab 551 \tab 884 \tab 10 \tab - \tab - \tab 2832 \cr
   2002 \tab 1393 \tab 580 \tab 778 \tab 6 \tab 8 \tab - \tab 2765 \cr
   Total \tab 2780 \tab 1131 \tab 1662 \tab 16 \tab 8 \tab 0 \tab 5597 \cr
}
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{bigband}},
\code{\link{blues}},
\code{\link{blugrass}},
\code{\link{camping}},
\code{\link{classicl}},
\code{\link{classics}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{conrock}},
\code{\link{country}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{english}},
\code{\link{excelart}},
\code{\link{folk}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{gospel}},
\code{\link{grtbooks}},
\code{\link{hosthome}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{huntfish}},
\code{\link{hvymetal}},
\code{\link{impcultr}},
\code{\link{impfinan}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{jazz}},
\code{\link{judgeart}},
\code{\link{latin}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{makeart}},
\code{\link{modpaint}},
\code{\link{moodeasy}},
\code{\link{mostimp1}},
\code{\link{mostimp2}},
\code{\link{mostimp3}},
\code{\link{mostimp4}},
\code{\link{musicals}},
\code{\link{newage}},
\code{\link{obtohelp}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{opoutcme}},
\code{\link{pclit}},
\code{\link{perform}},
\code{\link{plymusic}},
\code{\link{rap}},
\code{\link{readfict}},
\code{\link{reggae}},
\code{\link{richpoor}},
\code{\link{seemovie}},
\code{\link{selfirst}},
\code{\link{standup}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}},
\code{\link{united}},
\code{\link{usevcr}},
\code{\link{visitart}}

Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{beausprt}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{drama}},
\code{\link{effort}},
\code{\link{feelgod}},
\code{\link{fidgety}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godlove}},
\code{\link{godwatch}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{lesspain}},
\code{\link{localart}},
\code{\link{meditate}},
\code{\link{natarts}},
\code{\link{natlart}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{privpray}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{rellife}},
\code{\link{relpersn}},
\code{\link{sprtprsn}},
\code{\link{stateart}},
\code{\link{uniongod}},
\code{\link{volarts}},
\code{\link{wrthless}}

Other Leisure: 
\code{\link{artgn}},
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{buyinfgn}},
\code{\link{camping}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{gamegn}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{hrsrelax}},
\code{\link{huntfish}},
\code{\link{investgn}},
\code{\link{makeart}},
\code{\link{memlit}},
\code{\link{memsport}},
\code{\link{newjobgn}},
\code{\link{partpol}},
\code{\link{peoplegn}},
\code{\link{perform}},
\code{\link{perfrmgn}},
\code{\link{plymusic}},
\code{\link{polcangn}},
\code{\link{poldisgn}},
\code{\link{polinfgn}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{seemovie}},
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timepdwk}},
\code{\link{timerelx}},
\code{\link{travelgn}},
\code{\link{usevcr}},
\code{\link{visart}},
\code{\link{visitart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}},
\code{\link{volarts}}
}
\concept{Culture}
\concept{Leisure}
\concept{Religion and Culture}
\keyword{variable}
