% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_266.R
\name{porn30}
\alias{porn30}
\title{R use pornographic site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
porn30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
Q. Sexually explicit material?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} 1-2 times
\item \verb{[3]} 3-5 times
\item \verb{[4]} more than 5 times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2410/vshow}{GSS Data Explorer page for \code{porn30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 575 \tab 61 \tab 14 \tab 18 \tab 74 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 850 \tab 96 \tab 22 \tab 20 \tab 95 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 620 \tab 36 \tab 14 \tab 7 \tab 13 \tab - \tab 2812 \cr
   Total \tab 5878 \tab 2045 \tab 193 \tab 50 \tab 45 \tab 182 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{invest12}},
\code{\link{person30}},
\code{\link{sci30}},
\code{\link{work12}}
}
\concept{Information Society}
\keyword{variable}
