% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornfirm}
\alias{pornfirm}
\title{How firm is r's opinion on porn issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornfirm
}
\details{
Question C. How firm are you about (your opinion on/the way you feel about) pornography--would you way you are very likely to change (your opinion), somewhat likely to change, somewhat unlikely to change or very unlikely to change your opinion?
}
\section{Values}{

\itemize{
\item \code{1} change very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} change very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/645/vshow}{GSS Data Explorer page for \code{pornfirm}}.
}

\seealso{
Other Core: 
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}},
\code{\link{xmovie1}}

Other Sex: 
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{xmovie}},
\code{\link{xmovie1}}

Other Split Ballots: 
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{xmovie1}}
}
\concept{Core}
\concept{Sex}
\concept{Split Ballots}
\keyword{variable}
