% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornimp}
\alias{pornimp}
\title{Importance of pornography issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornimp
}
\details{
Question A. How important is the pornography issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most imp
\item \verb{[2]} important
\item \verb{[3]} not very imp
\item \verb{[4]} not imp at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/643/vshow}{GSS Data Explorer page for \code{pornimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not imp at all \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 161 \tab 481 \tab 577 \tab 223 \tab 17 \tab 14 \tab 1473 \cr
   Total \tab 0 \tab 161 \tab 481 \tab 577 \tab 223 \tab 17 \tab 14 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{pornfirm}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Core}
\keyword{variable}
