% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{porninf}
\alias{porninf}
\title{Materials provide info about sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
porninf
}
\details{
Question The next questions are about pornography--books, movies, magazines, and photographs that show or describe sex activities.
I''m going to read some opinions about the effects of looking at or reading such sexual materials. As I read each one, please tell me if you think sexual materials do or do not have that effect.
READ EACH ITEM. CODE ONE FOR EACH
A. Sexual materials provide information about sex
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/636/vshow}{GSS Data Explorer page for \code{porninf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 928 \tab 488 \tab 83 \tab 5 \tab 1504 \cr
   1975 \tab - \tab 925 \tab 434 \tab 127 \tab 4 \tab 1490 \cr
   1976 \tab - \tab 844 \tab 544 \tab 105 \tab 6 \tab 1499 \cr
   1978 \tab - \tab 939 \tab 498 \tab 92 \tab 3 \tab 1532 \cr
   1980 \tab - \tab 854 \tab 504 \tab 107 \tab 3 \tab 1468 \cr
   1983 \tab - \tab 953 \tab 559 \tab 79 \tab 8 \tab 1599 \cr
   1984 \tab - \tab 842 \tab 547 \tab 74 \tab 10 \tab 1473 \cr
   1986 \tab - \tab 836 \tab 555 \tab 75 \tab 4 \tab 1470 \cr
   1987 \tab - \tab 1124 \tab 568 \tab 118 \tab 9 \tab 1819 \cr
   1988 \tab 484 \tab 580 \tab 350 \tab 63 \tab 4 \tab 1481 \cr
   1989 \tab 502 \tab 586 \tab 392 \tab 52 \tab 5 \tab 1537 \cr
   1990 \tab 473 \tab 502 \tab 339 \tab 56 \tab 2 \tab 1372 \cr
   1991 \tab 500 \tab 563 \tab 386 \tab 66 \tab 2 \tab 1517 \cr
   1993 \tab 549 \tab 573 \tab 419 \tab 55 \tab 10 \tab 1606 \cr
   1994 \tab 1965 \tab 605 \tab 367 \tab 51 \tab 4 \tab 2992 \cr
   Total \tab 4473 \tab 11654 \tab 6950 \tab 1203 \tab 79 \tab 24359 \cr
}
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{premarsx}},
\code{\link{spdue}},
\code{\link{sppaid}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\keyword{variable}
