% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{porninfo}
\alias{porninfo}
\title{How much info does r have on porn issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
porninfo
}
\details{
Question B. How much information do you have about the pornography issue--do you have all of the information you need, most of the information, some of the information, or very little information?
}
\section{Values}{

\itemize{
\item \code{1} all info r needs
\item \code{2} most info
\item \code{3} some info
\item \code{4} very little info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/644/vshow}{GSS Data Explorer page for \code{porninfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all info r needs \tab most info \tab some info \tab very little info \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 209 \tab 303 \tab 542 \tab 381 \tab 18 \tab 20 \tab 1473 \cr
   Total \tab 0 \tab 209 \tab 303 \tab 542 \tab 381 \tab 18 \tab 20 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Core}
\keyword{variable}
