% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornlaw}
\alias{pornlaw}
\title{Feelings about pornography laws}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornlaw
}
\details{
Question Which of these statements comes closest to your feelings about pornography laws?
READ FIRST THREE CATEGORIES ONLY. CIRCLE ONLY ONE CODE.
}
\section{Values}{

\itemize{
\item \code{1} there should be laws against the distribution of pornography whatever the age
\item \code{2} there should be laws against the distribution of pornography to persons under 18
\item \code{3} there should be no laws forbidding the distribution of pornography
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
   1996 \tab B/C \tab partial \cr
   1998 \tab B/C \tab partial \cr
   2000 \tab B/C \tab partial \cr
   2002 \tab B/C \tab partial \cr
   2004 \tab B/C \tab partial \cr
   2006 \tab B/C \tab partial \cr
   2008 \tab B/C \tab partial \cr
   2010 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab B/C \tab partial \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/640/vshow}{GSS Data Explorer page for \code{pornlaw}}.
}

\seealso{
Other Core: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}},
\code{\link{xmovie1}}

Other Sex: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Core}
\concept{Sex}
\keyword{variable}
