% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornout}
\alias{pornout}
\title{Materials provide outlet}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornout
}
\details{
Question The next questions are about pornography--books, movies, magazines, and photographs that show or describe sex activities.
I''m going to read some opinions about the effects of looking at or reading such sexual materials. As I read each one, please tell me if you think sexual materials do or do not have that effect.
READ EACH ITEM. CODE ONE FOR EACH
D. Sexual materials provide an outlet for bottled-up impulses
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/639/vshow}{GSS Data Explorer page for \code{pornout}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 820 \tab 528 \tab 151 \tab 5 \tab 1504 \cr
   1975 \tab - \tab 836 \tab 413 \tab 236 \tab 5 \tab 1490 \cr
   1976 \tab - \tab 829 \tab 427 \tab 235 \tab 8 \tab 1499 \cr
   1978 \tab - \tab 899 \tab 445 \tab 183 \tab 5 \tab 1532 \cr
   1980 \tab - \tab 865 \tab 393 \tab 204 \tab 6 \tab 1468 \cr
   1983 \tab - \tab 911 \tab 501 \tab 176 \tab 11 \tab 1599 \cr
   1984 \tab - \tab 887 \tab 399 \tab 177 \tab 10 \tab 1473 \cr
   1986 \tab - \tab 870 \tab 447 \tab 145 \tab 8 \tab 1470 \cr
   1987 \tab - \tab 1075 \tab 479 \tab 253 \tab 12 \tab 1819 \cr
   1988 \tab 484 \tab 557 \tab 303 \tab 132 \tab 5 \tab 1481 \cr
   1989 \tab 502 \tab 617 \tab 271 \tab 141 \tab 6 \tab 1537 \cr
   1990 \tab 473 \tab 528 \tab 239 \tab 129 \tab 3 \tab 1372 \cr
   1991 \tab 500 \tab 567 \tab 307 \tab 139 \tab 4 \tab 1517 \cr
   1993 \tab 549 \tab 634 \tab 275 \tab 138 \tab 10 \tab 1606 \cr
   1994 \tab 1965 \tab 587 \tab 303 \tab 132 \tab 5 \tab 2992 \cr
   Total \tab 4473 \tab 11482 \tab 5730 \tab 2571 \tab 103 \tab 24359 \cr
}
}

\seealso{
Other Core: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Core}
\keyword{variable}
