% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{posslqy}
\alias{posslqy}
\title{Relationship status and cohabitation or not}
\source{
General Social Survey https://gss.norc.org
}
\description{
posslqy
}
\details{
Question Which of these statements applies to you?  CODE ONE ONLY.
}
\section{Values}{

\itemize{
\item \verb{[1]} i am married and living in the same household as my husband or wife
\item \verb{[2]} i have a steady partner, and we live in the same household
\item \verb{[3]} i have a husband or wife or steady partner, but we don't live in the same household
\item \verb{[4]} i don't have a steady partner
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4152/vshow}{GSS Data Explorer page for \code{posslqy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab i am married and living in the same household as my husband or wife \tab i have a steady partner, and we live in the same household \tab i have a husband or wife or steady partner, but we don't live in the same household \tab i don't have a steady partner \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2012 \tab 998 \tab 433 \tab 97 \tab 96 \tab 348 \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 562 \tab 120 \tab 117 \tab 462 \tab 1 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 1442 \tab 576 \tab 149 \tab 130 \tab 570 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1189 \tab 467 \tab 128 \tab 103 \tab 460 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 990 \tab 213 \tab 153 \tab 684 \tab - \tab - \tab 5 \tab 4032 \cr
   2022 \tab 1780 \tab 693 \tab 179 \tab 143 \tab 733 \tab 8 \tab 3 \tab 5 \tab 3544 \cr
   2024 \tab 1683 \tab 653 \tab 146 \tab 136 \tab 683 \tab - \tab 5 \tab 3 \tab 3309 \cr
   Total \tab 10353 \tab 4374 \tab 1032 \tab 878 \tab 3940 \tab 12 \tab 10 \tab 13 \tab 20612 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{bestfrd}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{comyear}},
\code{\link{cousins}},
\code{\link{cowrkfrd}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{hapunhap}},
\code{\link{helpfrds}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{inlawnum}},
\code{\link{kidcall}},
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidpars}},
\code{\link{kidvisit}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{localgvt}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{malive}},
\code{\link{marcohab}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{neifrd}},
\code{\link{niecenep}},
\code{\link{othfrd}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{palive}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{peocntct}},
\code{\link{posslq}},
\code{\link{relcall}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{talkedto}},
\code{\link{trustpeo}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{usefrds}},
\code{\link{wantbest}}

Other Marital Status: 
\code{\link{agewed}},
\code{\link{agewedcr}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{evdiv}},
\code{\link{hapmar}},
\code{\link{marital}},
\code{\link{marnum}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{spdue}},
\code{\link{spmarnum}},
\code{\link{widowed}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{Marital Status}
\keyword{variable}
