% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_36.R
\name{postlf1}
\alias{postlf1}
\title{Life of peace and tranquility}
\source{
General Social Survey https://gss.norc.org
}
\description{
postlf1
}
\details{
Question Do you believe there is a life after death?
A. Of course, no one knows exactly what life after death would be like, but here are some ideas people have had. (HAND CARD AA) How likely do you feel each possibility is? Would you say very likely, somewhat likely, not too likely, or not likely at all? (CIRCLE ONE NUMBER BESIDE EACH IDEA)
\enumerate{
\item A life of peace and tranquility.
}
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not too likely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/301/vshow}{GSS Data Explorer page for \code{postlf1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not too likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab 389 \tab 770 \tab 346 \tab 39 \tab 9 \tab 38 \tab 8 \tab 1599 \cr
   1984 \tab 278 \tab 775 \tab 307 \tab 50 \tab 17 \tab 31 \tab 15 \tab 1473 \cr
   Total \tab 667 \tab 1545 \tab 653 \tab 89 \tab 26 \tab 69 \tab 23 \tab 3072 \cr
}
}

\seealso{
Other Core: 
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}}
}
\concept{Core}
\keyword{variable}
