% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_36.R
\name{postlf6}
\alias{postlf6}
\title{A spiritual life involving mind not body}
\source{
General Social Survey https://gss.norc.org
}
\description{
postlf6
}
\details{
Question Do you believe there is a life after death?
A. Of course, no one knows exactly what life after death would be like, but here are some ideas people have had. (HAND CARD AA) How likely do you feel each possibility is? Would you say very likely, somewhat likely, not too likely, or not likely at all? (CIRCLE ONE NUMBER BESIDE EACH IDEA)
6. A spiritual life, involving our mind but not our body.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not too likely
\item \code{4} not likely at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/306/vshow}{GSS Data Explorer page for \code{postlf6}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not too likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab 389 \tab 556 \tab 356 \tab 97 \tab 126 \tab 63 \tab 12 \tab 1599 \cr
   1984 \tab 278 \tab 542 \tab 336 \tab 109 \tab 128 \tab 61 \tab 19 \tab 1473 \cr
   Total \tab 667 \tab 1098 \tab 692 \tab 206 \tab 254 \tab 124 \tab 31 \tab 3072 \cr
}
}

\seealso{
Other Core: 
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}}
}
\concept{Core}
\keyword{variable}
