% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{povline}
\alias{povline}
\title{R's hh poverty status}
\source{
General Social Survey https://gss.norc.org
}
\description{
povline
}
\details{
Question A. Poverty Status of Household
}
\section{Values}{

\itemize{
\item \verb{[1]} poor
\item \verb{[2]} poor -- imputed
\item \verb{[3]} not poor
\item \verb{[4]} not poor -- imputed
\item \verb{[5]} borderline
\item \verb{[6]} borderline -- imputed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5196/vshow}{GSS Data Explorer page for \code{povline}}.

Counts by year:\tabular{lllllllllll}{
   year \tab poor \tab poor -- imputed \tab not poor \tab not poor -- imputed \tab borderline \tab borderline -- imputed \tab not imputable \tab not available in this year \tab iap \tab Total \cr
   1972 \tab 145 \tab 40 \tab 1155 \tab 81 \tab 168 \tab 6 \tab 18 \tab - \tab - \tab 1613 \cr
   1973 \tab 106 \tab 31 \tab 1174 \tab 69 \tab 104 \tab 3 \tab 17 \tab - \tab - \tab 1504 \cr
   1974 \tab 99 \tab 29 \tab 1158 \tab 83 \tab 100 \tab 5 \tab 10 \tab - \tab - \tab 1484 \cr
   1975 \tab 107 \tab 24 \tab 1185 \tab 48 \tab 117 \tab 1 \tab 8 \tab - \tab - \tab 1490 \cr
   1976 \tab 114 \tab 28 \tab 1151 \tab 69 \tab 126 \tab 4 \tab 7 \tab - \tab - \tab 1499 \cr
   1977 \tab 79 \tab 29 \tab 1246 \tab 86 \tab 74 \tab 3 \tab 13 \tab - \tab - \tab 1530 \cr
   1978 \tab 143 \tab 18 \tab 1203 \tab 75 \tab 83 \tab 1 \tab 9 \tab - \tab - \tab 1532 \cr
   1980 \tab 130 \tab 12 \tab 1169 \tab 81 \tab 63 \tab - \tab 13 \tab - \tab - \tab 1468 \cr
   1982 \tab 286 \tab 18 \tab 1310 \tab 131 \tab 87 \tab 3 \tab 25 \tab - \tab - \tab 1860 \cr
   1983 \tab 185 \tab 9 \tab 1205 \tab 127 \tab 59 \tab - \tab 14 \tab - \tab - \tab 1599 \cr
   1984 \tab 183 \tab 14 \tab 1105 \tab 99 \tab 58 \tab 2 \tab 12 \tab - \tab - \tab 1473 \cr
   1985 \tab 169 \tab 6 \tab 1182 \tab 100 \tab 63 \tab - \tab 14 \tab - \tab - \tab 1534 \cr
   1986 \tab 190 \tab 7 \tab 1096 \tab 108 \tab 54 \tab 2 \tab 13 \tab - \tab - \tab 1470 \cr
   1987 \tab 271 \tab 10 \tab 1346 \tab 126 \tab 45 \tab 1 \tab 20 \tab - \tab - \tab 1819 \cr
   1988 \tab 167 \tab 1 \tab 1142 \tab 111 \tab 55 \tab - \tab 5 \tab - \tab - \tab 1481 \cr
   1989 \tab 142 \tab 5 \tab 1204 \tab 139 \tab 37 \tab 1 \tab 9 \tab - \tab - \tab 1537 \cr
   1990 \tab 134 \tab 1 \tab 1059 \tab 125 \tab 44 \tab - \tab 9 \tab - \tab - \tab 1372 \cr
   1991 \tab 196 \tab 4 \tab 1125 \tab 131 \tab 47 \tab 1 \tab 13 \tab - \tab - \tab 1517 \cr
   1993 \tab 186 \tab 3 \tab 1231 \tab 113 \tab 42 \tab - \tab 31 \tab - \tab - \tab 1606 \cr
   Total \tab 3032 \tab 289 \tab 22446 \tab 1902 \tab 1426 \tab 33 \tab 260 \tab 0 \tab 0 \tab 29388 \cr
}
}

\seealso{
Other Income: 
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{incdef}},
\code{\link{minfour}},
\code{\link{realinc}},
\code{\link{realrinc}}
}
\concept{Income}
\keyword{variable}
