% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{powrorgs}
\alias{powrorgs}
\title{Intl orgs take away much power from american govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
powrorgs
}
\details{
Question How much do you agree or disagree with the following statements? (Please, select one response)
D. International organizations are taking away too much power from the American government.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2021 \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4886/vshow}{GSS Data Explorer page for \code{powrorgs}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2004 \tab 1596 \tab 112 \tab 303 \tab 377 \tab 329 \tab 28 \tab 66 \tab 1 \tab - \tab 2812 \cr
   2014 \tab 1264 \tab 75 \tab 322 \tab 351 \tab 362 \tab 35 \tab 125 \tab 4 \tab - \tab 2538 \cr
   2021 \tab 2003 \tab 174 \tab 377 \tab 618 \tab 446 \tab 131 \tab 112 \tab 154 \tab 17 \tab 4032 \cr
   Total \tab 4863 \tab 361 \tab 1002 \tab 1346 \tab 1137 \tab 194 \tab 303 \tab 159 \tab 17 \tab 9382 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
