% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{prayer}
\alias{prayer}
\title{Bible prayer in public schools}
\source{
General Social Survey https://gss.norc.org
}
\description{
prayer
}
\details{
Question A. The United States Supreme Court has ruled that no state or local government may require the reading of the Lord's Prayer or Bible verses in public schools. What are your views on this--do you approve or disapprove of the court ruling?
}
\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
   2004 \tab A/B \tab partial \cr
   2006 \tab A/B \tab partial \cr
   2008 \tab A/B \tab partial \cr
   2010 \tab A/B \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/361/vshow}{GSS Data Explorer page for \code{prayer}}.
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
