% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{prayerv}
\alias{prayerv}
\title{Bible prayer in public schools (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
prayerv
}
\details{
Question The United States Supreme Court has ruled that no state or local government may require the reading of the Lord's Prayer or Bible verses in public schools. What are your views on this--do you approve or disapprove of the court ruling?
}
\section{Values}{

\itemize{
\item \verb{[1]} approve
\item \verb{[2]} disapprove
\item \verb{[3]} no opinion (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7585/vshow}{GSS Data Explorer page for \code{prayerv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab approve \tab disapprove \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2658 \tab 819 \tab 457 \tab 93 \tab 5 \tab 4032 \cr
   2022 \tab 2988 \tab 299 \tab 146 \tab 111 \tab - \tab 3544 \cr
   Total \tab 5646 \tab 1118 \tab 603 \tab 204 \tab 5 \tab 7576 \cr
}
}

\keyword{variable}
