% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{prayerx}
\alias{prayerx}
\title{Bible prayer in schools:version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
prayerx
}
\details{
Question C. Which of these statements comes closest to your feelings about prayers in the public schools?
The Lord''s prayer or some Bible verse should be read daily.
There should be a moment for silent prayer or meditation daily.
No prayer or other religious observances should be held in the public schools.
Other (SPECIFY).
}
\section{Values}{

\itemize{
\item \verb{[1]} prayer read daily
\item \verb{[2]} silent meditation
\item \verb{[3]} no prayers
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/363/vshow}{GSS Data Explorer page for \code{prayerx}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab prayer read daily \tab silent meditation \tab no prayers \tab other \tab don't know \tab no answer \tab Total \cr
   1986 \tab 730 \tab 209 \tab 382 \tab 126 \tab 11 \tab 5 \tab 7 \tab 1470 \cr
   Total \tab 730 \tab 209 \tab 382 \tab 126 \tab 11 \tab 5 \tab 7 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayery}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
