% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{prayery}
\alias{prayery}
\title{Bible prayer in schools-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
prayery
}
\details{
Question B. What are your views on the reading of the Lord's Prayer or Bible verses in public schools? Do you think it should be required in all public schools, not allowed in any public schools, or that it should be up to each state or local community to decide?
}
\section{Values}{

\itemize{
\item \verb{[1]} required
\item \verb{[2]} forbidden
\item \verb{[3]} local decision
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/362/vshow}{GSS Data Explorer page for \code{prayery}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab required \tab forbidden \tab local decision \tab don't know \tab no answer \tab Total \cr
   1974 \tab 753 \tab 224 \tab 60 \tab 420 \tab 19 \tab 8 \tab 1484 \cr
   Total \tab 753 \tab 224 \tab 60 \tab 420 \tab 19 \tab 8 \tab 1484 \cr
}
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
