% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{preach}
\alias{preach}
\title{How well local church preaches}
\source{
General Social Survey https://gss.norc.org
}
\description{
preach
}
\details{
Question How do you rate how your local church does each of the following?  Would you say your local church does an excellent, very good, fair, or poor job of:
A. Preaching
}
\section{Values}{

\itemize{
\item \verb{[1]} excellent
\item \verb{[2]} very good
\item \verb{[3]} good
\item \verb{[4]} fair
\item \verb{[5]} poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1161/vshow}{GSS Data Explorer page for \code{preach}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1831 \tab 260 \tab 280 \tab 243 \tab 61 \tab 18 \tab 51 \tab 73 \tab 2817 \cr
   Total \tab 1831 \tab 260 \tab 280 \tab 243 \tab 61 \tab 18 \tab 51 \tab 73 \tab 2817 \cr
}
}

\seealso{
Other Religion: 
\code{\link{ferespct}},
\code{\link{obeytch}},
\code{\link{sympcoun}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
