% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{prefsets}
\alias{prefsets}
\title{How familiar r is with term preference settings}
\source{
General Social Survey https://gss.norc.org
}
\description{
prefsets
}
\details{
Question H. Preference settings
}
\section{Values}{

\itemize{
\item \code{1} very familiar
\item \code{2} somewhat familiar
\item \code{3} not familiar
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2637/vshow}{GSS Data Explorer page for \code{prefsets}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab very familiar \tab somewhat familiar \tab not familiar \tab no answer \tab Total \cr
   2002 \tab 2142 \tab 221 \tab 201 \tab 154 \tab 47 \tab 2765 \cr
   2004 \tab 992 \tab 774 \tab 598 \tab 429 \tab 19 \tab 2812 \cr
   Total \tab 3134 \tab 995 \tab 799 \tab 583 \tab 66 \tab 5577 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{browser3}},
\code{\link{ezines}},
\code{\link{mp3}},
\code{\link{newsgrps}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}}
}
\concept{Information Society}
\keyword{variable}
