% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{premarsx}
\alias{premarsx}
\title{Sex before marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
premarsx
}
\details{
Question There's been a lot of discussion about the way morals and attitudes about sex are changing in this country. If a man and woman have sex relations before marriage, do you think it is always wrong, almost always wrong, wrong only
sometimes, or not wrong at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} always wrong
\item \verb{[2]} almost always wrong
\item \verb{[3]} wrong only sometimes
\item \verb{[4]} not wrong at all
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1974, 1975, 1977, 1978, 1982, 1983, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/631/vshow}{GSS Data Explorer page for \code{premarsx}}.

Counts by year:\tabular{llllllllll}{
   year \tab always wrong \tab almost always wrong \tab wrong only sometimes \tab not wrong at all \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 563 \tab 182 \tab 373 \tab 419 \tab 65 \tab 11 \tab - \tab - \tab 1613 \cr
   1974 \tab 472 \tab 181 \tab 337 \tab 439 \tab 48 \tab 7 \tab - \tab - \tab 1484 \cr
   1975 \tab 441 \tab 176 \tab 342 \tab 468 \tab 58 \tab 5 \tab - \tab - \tab 1490 \cr
   1977 \tab 459 \tab 141 \tab 340 \tab 541 \tab 39 \tab 10 \tab - \tab - \tab 1530 \cr
   1978 \tab 437 \tab 175 \tab 304 \tab 578 \tab 34 \tab 4 \tab - \tab - \tab 1532 \cr
   1982 \tab 520 \tab 152 \tab 359 \tab 763 \tab 55 \tab 11 \tab - \tab - \tab 1860 \cr
   1983 \tab 427 \tab 155 \tab 378 \tab 601 \tab 31 \tab 7 \tab - \tab - \tab 1599 \cr
   1985 \tab 422 \tab 123 \tab 294 \tab 643 \tab 48 \tab 4 \tab - \tab - \tab 1534 \cr
   1986 \tab 396 \tab 127 \tab 328 \tab 574 \tab 39 \tab 6 \tab - \tab - \tab 1470 \cr
   1988 \tab 251 \tab 102 \tab 214 \tab 388 \tab 27 \tab 6 \tab 493 \tab - \tab 1481 \cr
   1989 \tab 272 \tab 84 \tab 221 \tab 394 \tab 31 \tab 4 \tab 531 \tab - \tab 1537 \cr
   1990 \tab 229 \tab 103 \tab 205 \tab 356 \tab 32 \tab 3 \tab 444 \tab - \tab 1372 \cr
   1991 \tab 274 \tab 98 \tab 186 \tab 435 \tab 28 \tab 3 \tab 493 \tab - \tab 1517 \cr
   1993 \tab 283 \tab 106 \tab 218 \tab 443 \tab 27 \tab 3 \tab 526 \tab - \tab 1606 \cr
   1994 \tab 496 \tab 192 \tab 389 \tab 830 \tab 60 \tab 10 \tab 1015 \tab - \tab 2992 \cr
   1996 \tab 452 \tab 183 \tab 429 \tab 832 \tab 58 \tab 6 \tab 944 \tab - \tab 2904 \cr
   1998 \tab 476 \tab 166 \tab 378 \tab 784 \tab 65 \tab 2 \tab 961 \tab - \tab 2832 \cr
   2000 \tab 502 \tab 157 \tab 384 \tab 749 \tab 75 \tab 10 \tab 940 \tab - \tab 2817 \cr
   2002 \tab 242 \tab 73 \tab 176 \tab 394 \tab 18 \tab 5 \tab 1857 \tab - \tab 2765 \cr
   2004 \tab 238 \tab 79 \tab 157 \tab 409 \tab 18 \tab 5 \tab 1906 \tab - \tab 2812 \cr
   2006 \tab 509 \tab 167 \tab 379 \tab 882 \tab 44 \tab 11 \tab 2518 \tab - \tab 4510 \cr
   2008 \tab 305 \tab 90 \tab 206 \tab 692 \tab 30 \tab 6 \tab 694 \tab - \tab 2023 \cr
   2010 \tab 305 \tab 116 \tab 241 \tab 742 \tab 21 \tab 5 \tab 614 \tab - \tab 2044 \cr
   2012 \tab 273 \tab 72 \tab 206 \tab 704 \tab 33 \tab 14 \tab 672 \tab - \tab 1974 \cr
   2014 \tab 324 \tab 111 \tab 257 \tab 956 \tab 18 \tab 9 \tab 863 \tab - \tab 2538 \cr
   2016 \tab 389 \tab 100 \tab 263 \tab 1101 \tab 25 \tab 10 \tab 979 \tab - \tab 2867 \cr
   2018 \tab 278 \tab 95 \tab 213 \tab 943 \tab 19 \tab 11 \tab 789 \tab - \tab 2348 \cr
   2021 \tab 390 \tab 162 \tab 367 \tab 1761 \tab 9 \tab 2 \tab 1315 \tab 26 \tab 4032 \cr
   2022 \tab 308 \tab 104 \tab 305 \tab 1607 \tab 36 \tab 8 \tab 1168 \tab 8 \tab 3544 \cr
   Total \tab 10933 \tab 3772 \tab 8449 \tab 20428 \tab 1091 \tab 198 \tab 19722 \tab 34 \tab 64627 \cr
}
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{porninf}},
\code{\link{spdue}},
\code{\link{sppaid}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\keyword{variable}
