% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{pres00}
\alias{pres00}
\title{Vote for gore, bush, nader}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres00
}
\details{
Question I. In 2000, you remember that Gore ran for President on the Democratic ticket against Bush for the Republicans . Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Gore or Bush?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} gore
\item \verb{[2]} bush
\item \verb{[3]} nader
\item \verb{[4]} other (specify)
\item \verb{[6]} didnt vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/167/vshow}{GSS Data Explorer page for \code{pres00}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab gore \tab bush \tab nader \tab other (specify) \tab didnt vote \tab don't know \tab no answer \tab Total \cr
   2002 \tab 985 \tab 781 \tab 885 \tab 57 \tab 16 \tab 10 \tab 6 \tab 25 \tab 2765 \cr
   2004 \tab 1048 \tab 738 \tab 922 \tab 28 \tab 2 \tab 4 \tab 21 \tab 49 \tab 2812 \cr
   2006 \tab 2674 \tab 813 \tab 903 \tab 26 \tab 19 \tab 9 \tab 14 \tab 52 \tab 4510 \cr
   Total \tab 4707 \tab 2332 \tab 2710 \tab 111 \tab 37 \tab 23 \tab 41 \tab 126 \tab 10087 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
