% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{pres04}
\alias{pres04}
\title{Vote for kerry, bush, nader}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres04
}
\details{
Question J. In 2004, you remember that Kerry ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Kerry or Bush?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} kerry
\item \verb{[2]} bush
\item \verb{[3]} nader
\item \verb{[4]} other (specify)
\item \verb{[6]} didnt vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/170/vshow}{GSS Data Explorer page for \code{pres04}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab kerry \tab bush \tab nader \tab didnt vote \tab don't know \tab no answer \tab Total \cr
   2006 \tab 1458 \tab 1434 \tab 1446 \tab 47 \tab 17 \tab 18 \tab 90 \tab 4510 \cr
   2008 \tab 768 \tab 580 \tab 599 \tab 25 \tab 12 \tab 8 \tab 31 \tab 2023 \cr
   2010 \tab 811 \tab 526 \tab 630 \tab 26 \tab 6 \tab 8 \tab 37 \tab 2044 \cr
   Total \tab 3037 \tab 2540 \tab 2675 \tab 98 \tab 35 \tab 34 \tab 158 \tab 8577 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
