% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{pres12}
\alias{pres12}
\title{Vote obama or romney}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres12
}
\details{
Question L. In 2012, you remember that Obama ran for President on the Democratic ticket against Romney for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item Did you vote for Obama or Romney?
}
}
\section{Values}{

\itemize{
\item \code{1} obama
\item \code{2} romney
\item \code{3} other candidate (specify)
\item \code{4} didn't vote for president
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/176/vshow}{GSS Data Explorer page for \code{pres12}}.
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
