% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{pres16}
\alias{pres16}
\title{Vote clinton or trump}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres16
}
\details{
Question Did you vote for Clinton or Trump?
}
\section{Values}{

\itemize{
\item \code{1} clinton
\item \code{2} trump
\item \code{3} other candidate
\item \code{4} didn't vote for president
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7090/vshow}{GSS Data Explorer page for \code{pres16}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab clinton \tab trump \tab other candidate \tab didn't vote for president \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 764 \tab 577 \tab 87 \tab 20 \tab 35 \tab 811 \tab 54 \tab - \tab 2348 \cr
   2021 \tab - \tab 1509 \tab 1037 \tab 169 \tab 49 \tab 77 \tab 902 \tab 250 \tab 39 \tab 4032 \cr
   2022 \tab - \tab 1117 \tab 830 \tab 144 \tab 21 \tab 63 \tab 1211 \tab 143 \tab 15 \tab 3544 \cr
   Total \tab 0 \tab 3390 \tab 2444 \tab 400 \tab 90 \tab 175 \tab 2924 \tab 447 \tab 54 \tab 9924 \cr
}
}

\keyword{variable}
