% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{pres20}
\alias{pres20}
\title{Voted trump or biden}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres20
}
\details{
Question Did you vote for Joe Biden or Donald Trump?
}
\section{Values}{

\itemize{
\item \verb{[1]} biden
\item \verb{[2]} trump
\item \verb{[3]} other candidate
\item \verb{[4]} didn't vote for president
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7759/vshow}{GSS Data Explorer page for \code{pres20}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab biden \tab trump \tab other candidate \tab didn't vote for president \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 1417 \tab 883 \tab 82 \tab 14 \tab 38 \tab 939 \tab 158 \tab 13 \tab 3544 \cr
   2024 \tab - \tab 1188 \tab 874 \tab 53 \tab 13 \tab 44 \tab 981 \tab 38 \tab 118 \tab 3309 \cr
   Total \tab 0 \tab 2605 \tab 1757 \tab 135 \tab 27 \tab 82 \tab 1920 \tab 196 \tab 131 \tab 6853 \cr
}
}

\keyword{variable}
