% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{pres68}
\alias{pres68}
\title{Vote for humphrey, nixon, or wallace}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres68
}
\details{
Question A. Now in 1968, you remember that Humphrey ran for President on the Democratic ticket against Nixon for the Republicans, and Wallace as an Independent. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Humphrey, Nixon, or Wallace?
}
}
\section{Values}{

\itemize{
\item \code{1} humphrey
\item \code{2} nixon
\item \code{3} wallace
\item \code{4} other
\item \code{5} refused
\item \code{6} wldnt vt-relig
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/143/vshow}{GSS Data Explorer page for \code{pres68}}.

Counts by year:\tabular{llllllllll}{
   year \tab humphrey \tab nixon \tab wallace \tab other \tab refused \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 429 \tab 478 \tab 93 \tab 9 \tab 21 \tab 26 \tab 551 \tab 6 \tab 1613 \cr
   1973 \tab 358 \tab 407 \tab 127 \tab 7 \tab 11 \tab 17 \tab 574 \tab 3 \tab 1504 \cr
   Total \tab 787 \tab 885 \tab 220 \tab 16 \tab 32 \tab 43 \tab 1125 \tab 9 \tab 3117 \cr
}
}

\seealso{
Other Core: 
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{partyid}},
\code{\link{spdotpre}},
\code{\link{vote68}}
}
\concept{Core}
\keyword{variable}
