% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{pres76}
\alias{pres76}
\title{Vote for carter or ford}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres76
}
\details{
Question C. In 1976, you remember that Carter ran for President on the Democratic ticket against Ford for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Carter or Ford?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} carter
\item \verb{[2]} ford
\item \verb{[3]} other
\item \verb{[4]} refused
\item \verb{[5]} no pres. vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1978, 1980, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/149/vshow}{GSS Data Explorer page for \code{pres76}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab carter \tab ford \tab other \tab no pres. vote \tab don't know \tab no answer \tab refused \tab Total \cr
   1977 \tab 542 \tab 528 \tab 425 \tab 9 \tab 3 \tab 7 \tab 16 \tab - \tab 1530 \cr
   1978 \tab 557 \tab 497 \tab 428 \tab 15 \tab - \tab 14 \tab 17 \tab 4 \tab 1532 \cr
   1980 \tab 569 \tab 495 \tab 359 \tab 9 \tab 5 \tab 13 \tab 7 \tab 11 \tab 1468 \cr
   1982 \tab 757 \tab 694 \tab 359 \tab 5 \tab - \tab 28 \tab 6 \tab 11 \tab 1860 \cr
   Total \tab 2425 \tab 2214 \tab 1571 \tab 38 \tab 8 \tab 62 \tab 46 \tab 26 \tab 6390 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
