% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{pres80}
\alias{pres80}
\title{Vote for carter, reagan or anderson}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres80
}
\details{
Question D. In 1980, you remember that Carter ran for President on the Democratic ticket against Reagan for the Republicans, and  Anderson as an Independent. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:
Did you vote for Carter, Reagan, or Anderson?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} carter
\item \verb{[2]} reagan
\item \verb{[3]} anderson
\item \verb{[4]} other
\item \verb{[5]} refused
\item \verb{[6]} didnt vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984, 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/152/vshow}{GSS Data Explorer page for \code{pres80}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab carter \tab reagan \tab anderson \tab other \tab refused \tab didnt vote \tab don't know \tab no answer \tab Total \cr
   1982 \tab 674 \tab 632 \tab 442 \tab 70 \tab 4 \tab 17 \tab 5 \tab 13 \tab 3 \tab 1860 \cr
   1983 \tab 531 \tab 482 \tab 470 \tab 71 \tab 4 \tab 12 \tab 3 \tab 20 \tab 6 \tab 1599 \cr
   1984 \tab 519 \tab 418 \tab 419 \tab 67 \tab 7 \tab 12 \tab 2 \tab 19 \tab 10 \tab 1473 \cr
   1985 \tab 513 \tab 469 \tab 472 \tab 46 \tab 6 \tab 11 \tab 3 \tab 12 \tab 2 \tab 1534 \cr
   1987 \tab 671 \tab 634 \tab 415 \tab 54 \tab 4 \tab - \tab - \tab 18 \tab 23 \tab 1819 \cr
   Total \tab 2908 \tab 2635 \tab 2218 \tab 308 \tab 25 \tab 52 \tab 13 \tab 82 \tab 44 \tab 8285 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
