% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{pres84}
\alias{pres84}
\title{Vote for reagan or mondale}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres84
}
\details{
Question E. In 1984, you remember that Mondale ran for President on the Democratic ticket against Reagan for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Mondale or Reagan?
}
}
\section{Values}{

\itemize{
\item \code{1} mondale
\item \code{2} reagan
\item \code{3} other
\item \code{4} refused
\item \code{5} no pres. vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989 \cr
   Full \tab 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/155/vshow}{GSS Data Explorer page for \code{pres84}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab mondale \tab reagan \tab other \tab refused \tab no pres. vote \tab don't know \tab no answer \tab Total \cr
   1985 \tab 475 \tab 399 \tab 618 \tab 4 \tab 19 \tab 10 \tab 7 \tab 2 \tab 1534 \cr
   1986 \tab 487 \tab 372 \tab 585 \tab 8 \tab 7 \tab 3 \tab 8 \tab - \tab 1470 \cr
   1987 \tab 642 \tab 550 \tab 566 \tab 7 \tab 5 \tab - \tab 22 \tab 27 \tab 1819 \cr
   1988 \tab 572 \tab 310 \tab 546 \tab 11 \tab 7 \tab 5 \tab 23 \tab 7 \tab 1481 \cr
   1989 \tab 555 \tab 330 \tab 596 \tab 7 \tab - \tab 3 \tab 21 \tab 25 \tab 1537 \cr
   Total \tab 2731 \tab 1961 \tab 2911 \tab 37 \tab 38 \tab 21 \tab 81 \tab 61 \tab 7841 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
