% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{pres88}
\alias{pres88}
\title{Vote for dukakis or bush}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres88
}
\details{
Question F. In 1988, you remember that Dukakis ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Dukakis or Bush?
}
}
\section{Values}{

\itemize{
\item \code{1} dukakis
\item \code{2} bush
\item \code{3} other
\item \code{4} refused
\item \code{5} no pres. vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/158/vshow}{GSS Data Explorer page for \code{pres88}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab dukakis \tab bush \tab other \tab refused \tab no pres. vote \tab don't know \tab no answer \tab Total \cr
   1989 \tab 551 \tab 431 \tab 505 \tab 8 \tab 1 \tab 3 \tab 12 \tab 26 \tab 1537 \cr
   1990 \tab 483 \tab 312 \tab 529 \tab 11 \tab 1 \tab - \tab 19 \tab 17 \tab 1372 \cr
   1991 \tab 532 \tab 381 \tab 557 \tab 9 \tab - \tab 7 \tab 20 \tab 11 \tab 1517 \cr
   1993 \tab 593 \tab 282 \tab 686 \tab 10 \tab - \tab 4 \tab 15 \tab 16 \tab 1606 \cr
   Total \tab 2159 \tab 1406 \tab 2277 \tab 38 \tab 2 \tab 14 \tab 66 \tab 70 \tab 6032 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
