% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{pres92}
\alias{pres92}
\title{Vote for clinton, bush, perot}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres92
}
\details{
Question G. In 1992, you remember that Clinton ran for President on the Democratic ticket against Bush for the Republicans and Perot as an Independent.
Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:
Did you vote for Clinton, Bush, or Perot?
}
}
\section{Values}{

\itemize{
\item \code{1} clinton
\item \code{2} bush
\item \code{3} perot
\item \code{4} other
\item \code{6} no pres. vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/161/vshow}{GSS Data Explorer page for \code{pres92}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab clinton \tab bush \tab perot \tab other \tab no pres. vote \tab don't know \tab no answer \tab Total \cr
   1993 \tab 493 \tab 464 \tab 416 \tab 200 \tab 4 \tab 2 \tab 8 \tab 19 \tab 1606 \cr
   1994 \tab 938 \tab 940 \tab 699 \tab 349 \tab 5 \tab 6 \tab 23 \tab 32 \tab 2992 \cr
   1996 \tab 995 \tab 913 \tab 662 \tab 278 \tab 5 \tab - \tab 28 \tab 23 \tab 2904 \cr
   1998 \tab 1034 \tab 817 \tab 672 \tab 212 \tab 7 \tab 7 \tab 43 \tab 40 \tab 2832 \cr
   Total \tab 3460 \tab 3134 \tab 2449 \tab 1039 \tab 21 \tab 15 \tab 102 \tab 114 \tab 10334 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
