% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{pres96}
\alias{pres96}
\title{Vote for clinton, dole, perot}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres96
}
\details{
Question H. In 1996, you remember that Clinton ran for President on the Democratic ticket against Dole for the Republicans and Perot for the Reform Party. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Clinton, Bush, or Perot?
}
}
\section{Values}{

\itemize{
\item \code{1} clinton
\item \code{2} dole
\item \code{3} perot
\item \code{4} other (specify)
\item \code{6} didnt vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/164/vshow}{GSS Data Explorer page for \code{pres96}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab clinton \tab dole \tab perot \tab other (specify) \tab didnt vote \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1019 \tab 923 \tab 589 \tab 199 \tab 16 \tab 6 \tab 40 \tab 40 \tab 2832 \cr
   2000 \tab 1074 \tab 918 \tab 498 \tab 216 \tab 20 \tab 4 \tab 57 \tab 30 \tab 2817 \cr
   2002 \tab 1031 \tab 1008 \tab 475 \tab 197 \tab 18 \tab 10 \tab 7 \tab 19 \tab 2765 \cr
   Total \tab 3124 \tab 2849 \tab 1562 \tab 612 \tab 54 \tab 20 \tab 104 \tab 89 \tab 8414 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
