% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{prespop}
\alias{prespop}
\title{Approve of pres handling job}
\source{
General Social Survey https://gss.norc.org
}
\description{
prespop
}
\details{
Question Do you approve or disapprove of the way Barack Obama is handling his job as president?
}
\section{Values}{

\itemize{
\item \verb{[1]} approve
\item \verb{[2]} disapprove
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3969/vshow}{GSS Data Explorer page for \code{prespop}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab approve \tab disapprove \tab don't know \tab no answer \tab Total \cr
   2012 \tab 675 \tab 633 \tab 541 \tab 91 \tab 34 \tab 1974 \cr
   Total \tab 675 \tab 633 \tab 541 \tab 91 \tab 34 \tab 1974 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{volacty2}},
\code{\link{volactyr}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
