% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_01.R
\name{prestige}
\alias{prestige}
\title{R's occupational prestige score(1970)}
\source{
General Social Survey https://gss.norc.org
}
\description{
prestige
}
\details{
Question A. What kind of work do you (did you normally) do?  That is, what (is/was) your job called?
OCCUPATION:  (See REMARKS)
B. What (do/did) you actually do in that job?  Tell me, what (are/were) some of your main duties?
C. What kind of place (do/did) you work for?
INDUSTRY:  (See REMARKS)
D. What (do/did) they (make/do)?
E. (Are/Were) you self employed or (do/did) you work for someone else?
F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)?
G. About how much time (does/did) it usually take you to travel to work - about how many minutes?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 12 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 60 \tab 61 \tab 62 \tab 63 \tab 67 \tab 68 \tab 69 \tab 71 \tab 72 \tab 74 \tab 76 \tab 78 \tab 82 \tab dk, na, iap \tab 21 \tab 53 \tab 65 \tab 66 \tab 70 \tab not available in this year \tab Total \cr
   1972 \tab 12 \tab 11 \tab 5 \tab 18 \tab 41 \tab 62 \tab 18 \tab 32 \tab 20 \tab 9 \tab 1 \tab 35 \tab 30 \tab 18 \tab 8 \tab 46 \tab 17 \tab 23 \tab 93 \tab 33 \tab 43 \tab 13 \tab 120 \tab 17 \tab 13 \tab 24 \tab 40 \tab 74 \tab 24 \tab 22 \tab 12 \tab 39 \tab 66 \tab 38 \tab 54 \tab 16 \tab 93 \tab 25 \tab 6 \tab 5 \tab 2 \tab 11 \tab 14 \tab 4 \tab 45 \tab 16 \tab 26 \tab 21 \tab 2 \tab 3 \tab 6 \tab 3 \tab 4 \tab 3 \tab 1 \tab 9 \tab 1 \tab 166 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 9 \tab 8 \tab 3 \tab 22 \tab 40 \tab 37 \tab 15 \tab 32 \tab 33 \tab 14 \tab 3 \tab 32 \tab 28 \tab 22 \tab 8 \tab 55 \tab 7 \tab 27 \tab 82 \tab 28 \tab 33 \tab 10 \tab 104 \tab 17 \tab 4 \tab 22 \tab 42 \tab 54 \tab 11 \tab 18 \tab 15 \tab 33 \tab 82 \tab 37 \tab 61 \tab 9 \tab 93 \tab 14 \tab 9 \tab 3 \tab 2 \tab 9 \tab 15 \tab 1 \tab 27 \tab 14 \tab 24 \tab 20 \tab 5 \tab 4 \tab 4 \tab 2 \tab 7 \tab 2 \tab 5 \tab 12 \tab 1 \tab 172 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 8 \tab 8 \tab 6 \tab 16 \tab 45 \tab 41 \tab 9 \tab 28 \tab 28 \tab 16 \tab 3 \tab 27 \tab 33 \tab 21 \tab 10 \tab 61 \tab 6 \tab 28 \tab 77 \tab 30 \tab 11 \tab 5 \tab 116 \tab 18 \tab 7 \tab 23 \tab 45 \tab 56 \tab 13 \tab 15 \tab 5 \tab 40 \tab 74 \tab 37 \tab 61 \tab 16 \tab 76 \tab 18 \tab 6 \tab 5 \tab 6 \tab 18 \tab 6 \tab 4 \tab 62 \tab 11 \tab 24 \tab 26 \tab 2 \tab - \tab 13 \tab 3 \tab 8 \tab - \tab 5 \tab 12 \tab 2 \tab 133 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 8 \tab 9 \tab 7 \tab 18 \tab 27 \tab 41 \tab 18 \tab 41 \tab 19 \tab 22 \tab - \tab 50 \tab 29 \tab 22 \tab 21 \tab 26 \tab 12 \tab 25 \tab 69 \tab 31 \tab 58 \tab 9 \tab 95 \tab 24 \tab 8 \tab 25 \tab 39 \tab 50 \tab 24 \tab 20 \tab 6 \tab 35 \tab 73 \tab 33 \tab 67 \tab 10 \tab 94 \tab 20 \tab 5 \tab 2 \tab 4 \tab 6 \tab 15 \tab 7 \tab 35 \tab 20 \tab 20 \tab 14 \tab 4 \tab 1 \tab 13 \tab 1 \tab 9 \tab - \tab 3 \tab 4 \tab - \tab 141 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 8 \tab 5 \tab 7 \tab 13 \tab 39 \tab 56 \tab 15 \tab 29 \tab 23 \tab 6 \tab 4 \tab 49 \tab 34 \tab 26 \tab 9 \tab 28 \tab 14 \tab 28 \tab 68 \tab 30 \tab 68 \tab 15 \tab 97 \tab 24 \tab 5 \tab 27 \tab 38 \tab 68 \tab 14 \tab 15 \tab 15 \tab 37 \tab 70 \tab 40 \tab 54 \tab 9 \tab 83 \tab 12 \tab 7 \tab 2 \tab 7 \tab 11 \tab 11 \tab 1 \tab 32 \tab 9 \tab 19 \tab 26 \tab 2 \tab 5 \tab 9 \tab 3 \tab 8 \tab - \tab 4 \tab 7 \tab 2 \tab 147 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1499 \cr
   1977 \tab 5 \tab 12 \tab 7 \tab 16 \tab 50 \tab 41 \tab 27 \tab 32 \tab 27 \tab 21 \tab 1 \tab 33 \tab 30 \tab 31 \tab 19 \tab 33 \tab 21 \tab 30 \tab 85 \tab 40 \tab 62 \tab 12 \tab 100 \tab 29 \tab 7 \tab 24 \tab 49 \tab 40 \tab 11 \tab 16 \tab 11 \tab 31 \tab 76 \tab 32 \tab 38 \tab 8 \tab 86 \tab 14 \tab 6 \tab 5 \tab 4 \tab 10 \tab 18 \tab 4 \tab 43 \tab 20 \tab 21 \tab 28 \tab 5 \tab 4 \tab 14 \tab 4 \tab 5 \tab 3 \tab 1 \tab 10 \tab 2 \tab 114 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1530 \cr
   1978 \tab 10 \tab 6 \tab 12 \tab 12 \tab 30 \tab 44 \tab 22 \tab 40 \tab 19 \tab 20 \tab 2 \tab 38 \tab 34 \tab 21 \tab 12 \tab 42 \tab 8 \tab 22 \tab 66 \tab 37 \tab 69 \tab 11 \tab 77 \tab 19 \tab 5 \tab 34 \tab 43 \tab 46 \tab 20 \tab 10 \tab 17 \tab 31 \tab 87 \tab 52 \tab 66 \tab 14 \tab 85 \tab 14 \tab 10 \tab 3 \tab 4 \tab 12 \tab 13 \tab 2 \tab 52 \tab 27 \tab 22 \tab 17 \tab 4 \tab 3 \tab 8 \tab 4 \tab 11 \tab 2 \tab 3 \tab 6 \tab 4 \tab 128 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 11 \tab 7 \tab 1 \tab 21 \tab 32 \tab 31 \tab 11 \tab 40 \tab 20 \tab 15 \tab 3 \tab 40 \tab 38 \tab 19 \tab 14 \tab 33 \tab 10 \tab 12 \tab 54 \tab 28 \tab 51 \tab 15 \tab 100 \tab 18 \tab 7 \tab 30 \tab 36 \tab 52 \tab 20 \tab 13 \tab 17 \tab 35 \tab 72 \tab 41 \tab 61 \tab 12 \tab 106 \tab 34 \tab 8 \tab 7 \tab 5 \tab 20 \tab 21 \tab 4 \tab 38 \tab 12 \tab 18 \tab 17 \tab 5 \tab 1 \tab 5 \tab 1 \tab 5 \tab - \tab 8 \tab 7 \tab 3 \tab 119 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 40 \tab 12 \tab 4 \tab 13 \tab 63 \tab 79 \tab 29 \tab 31 \tab 48 \tab 22 \tab 2 \tab 39 \tab 45 \tab 23 \tab 14 \tab 38 \tab 15 \tab 30 \tab 99 \tab 44 \tab 67 \tab 16 \tab 191 \tab 33 \tab 6 \tab 22 \tab 41 \tab 47 \tab 19 \tab 21 \tab 7 \tab 47 \tab 73 \tab 39 \tab 55 \tab 14 \tab 98 \tab 24 \tab 14 \tab 5 \tab 3 \tab 17 \tab 9 \tab 2 \tab 57 \tab 18 \tab 27 \tab 24 \tab - \tab 1 \tab 4 \tab 3 \tab 7 \tab 1 \tab 11 \tab 7 \tab 1 \tab 137 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 4 \tab 5 \tab 9 \tab 15 \tab 32 \tab 21 \tab 15 \tab 36 \tab 26 \tab 24 \tab 2 \tab 47 \tab 39 \tab 28 \tab 13 \tab 37 \tab 11 \tab 39 \tab 46 \tab 25 \tab 61 \tab 16 \tab 101 \tab 22 \tab 9 \tab 18 \tab 37 \tab 73 \tab 12 \tab 9 \tab 9 \tab 39 \tab 106 \tab 47 \tab 73 \tab 13 \tab 129 \tab 25 \tab 13 \tab 1 \tab 5 \tab 5 \tab 20 \tab 3 \tab 41 \tab 30 \tab 24 \tab 22 \tab 5 \tab 5 \tab 9 \tab 1 \tab 9 \tab 4 \tab 4 \tab 9 \tab 2 \tab 106 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 1599 \cr
   1984 \tab 16 \tab 8 \tab 15 \tab 13 \tab 58 \tab 27 \tab 13 \tab 29 \tab 15 \tab 26 \tab 1 \tab 40 \tab 34 \tab 20 \tab 9 \tab 37 \tab 15 \tab 43 \tab 48 \tab 23 \tab 65 \tab 17 \tab 100 \tab 20 \tab 2 \tab 29 \tab 36 \tab 47 \tab 15 \tab 13 \tab 8 \tab 34 \tab 62 \tab 34 \tab 52 \tab 10 \tab 117 \tab 28 \tab 9 \tab 1 \tab 4 \tab 8 \tab 23 \tab 1 \tab 32 \tab 29 \tab 23 \tab 23 \tab 5 \tab 4 \tab 10 \tab 2 \tab 6 \tab - \tab 5 \tab 6 \tab 5 \tab 94 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 23 \tab 6 \tab 9 \tab 13 \tab 40 \tab 32 \tab 14 \tab 42 \tab 22 \tab 24 \tab 1 \tab 30 \tab 30 \tab 14 \tab 8 \tab 23 \tab 10 \tab 21 \tab 57 \tab 41 \tab 52 \tab 11 \tab 107 \tab 24 \tab 4 \tab 32 \tab 28 \tab 51 \tab 17 \tab 14 \tab 15 \tab 49 \tab 75 \tab 28 \tab 61 \tab 23 \tab 134 \tab 27 \tab 8 \tab 6 \tab 6 \tab 13 \tab 19 \tab 5 \tab 35 \tab 29 \tab 24 \tab 14 \tab 12 \tab 7 \tab 9 \tab 6 \tab 12 \tab 6 \tab 9 \tab 12 \tab 6 \tab 81 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1534 \cr
   1986 \tab 5 \tab 7 \tab 9 \tab 11 \tab 30 \tab 19 \tab 10 \tab 31 \tab 24 \tab 23 \tab 1 \tab 32 \tab 26 \tab 19 \tab 3 \tab 37 \tab 13 \tab 41 \tab 55 \tab 28 \tab 72 \tab 12 \tab 89 \tab 21 \tab 6 \tab 35 \tab 43 \tab 51 \tab 16 \tab 8 \tab 10 \tab 40 \tab 64 \tab 37 \tab 56 \tab 7 \tab 145 \tab 28 \tab 6 \tab 5 \tab 5 \tab 10 \tab 14 \tab 2 \tab 34 \tab 19 \tab 34 \tab 16 \tab 2 \tab 6 \tab 6 \tab 5 \tab 8 \tab 1 \tab 5 \tab 15 \tab 6 \tab 102 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1470 \cr
   1987 \tab 10 \tab 23 \tab 5 \tab 22 \tab 59 \tab 35 \tab 10 \tab 22 \tab 55 \tab 35 \tab 5 \tab 55 \tab 36 \tab 21 \tab 18 \tab 12 \tab 15 \tab 47 \tab 67 \tab 35 \tab 66 \tab 18 \tab 142 \tab 24 \tab 5 \tab 41 \tab 47 \tab 57 \tab 26 \tab 21 \tab 11 \tab 41 \tab 70 \tab 37 \tab 51 \tab 12 \tab 190 \tab 28 \tab 13 \tab 3 \tab 4 \tab 11 \tab 38 \tab 3 \tab 34 \tab 26 \tab 30 \tab 20 \tab 6 \tab 6 \tab 10 \tab 4 \tab 8 \tab 2 \tab 7 \tab 11 \tab 5 \tab 103 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 22 \tab 7 \tab 13 \tab 6 \tab 42 \tab 23 \tab 4 \tab 24 \tab 23 \tab 23 \tab 3 \tab 43 \tab 29 \tab 24 \tab 7 \tab 26 \tab 9 \tab 40 \tab 57 \tab 26 \tab 82 \tab 11 \tab 99 \tab 27 \tab 4 \tab 38 \tab 35 \tab 35 \tab 18 \tab 17 \tab 11 \tab 34 \tab 62 \tab 29 \tab 52 \tab 13 \tab 128 \tab 33 \tab 6 \tab 3 \tab 2 \tab 20 \tab 23 \tab 4 \tab 39 \tab 20 \tab 26 \tab 8 \tab 7 \tab 6 \tab 12 \tab 4 \tab 8 \tab 2 \tab 6 \tab 10 \tab 6 \tab 87 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 16 \tab 8 \tab 6 \tab 14 \tab 36 \tab 25 \tab 8 \tab 23 \tab 26 \tab 19 \tab 2 \tab 22 \tab 24 \tab 15 \tab 16 \tab 21 \tab 11 \tab 36 \tab 66 \tab 26 \tab 64 \tab 13 \tab 126 \tab 32 \tab 14 \tab 32 \tab 32 \tab 44 \tab 17 \tab 15 \tab 8 \tab 26 \tab 70 \tab 33 \tab 48 \tab 15 \tab 140 \tab 36 \tab 9 \tab 5 \tab 6 \tab 13 \tab 21 \tab 2 \tab 38 \tab 35 \tab 32 \tab 16 \tab 4 \tab 9 \tab 9 \tab 3 \tab 13 \tab 2 \tab 16 \tab 20 \tab 3 \tab 96 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 9 \tab 5 \tab 12 \tab 13 \tab 35 \tab 20 \tab 15 \tab 24 \tab 19 \tab 17 \tab 1 \tab 28 \tab 25 \tab 17 \tab 5 \tab 29 \tab 7 \tab 27 \tab 80 \tab 26 \tab 55 \tab 5 \tab 111 \tab 20 \tab 2 \tab 37 \tab 26 \tab 29 \tab 17 \tab 18 \tab 18 \tab 30 \tab 63 \tab 38 \tab 44 \tab 7 \tab 116 \tab 21 \tab 14 \tab 9 \tab 6 \tab 14 \tab 27 \tab 2 \tab 39 \tab 11 \tab 28 \tab 10 \tab 1 \tab 2 \tab 11 \tab 9 \tab 9 \tab 1 \tab 8 \tab 19 \tab 4 \tab 72 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 216 \tab 147 \tab 130 \tab 256 \tab 699 \tab 634 \tab 253 \tab 536 \tab 447 \tab 336 \tab 35 \tab 640 \tab 544 \tab 361 \tab 194 \tab 584 \tab 201 \tab 519 \tab 1169 \tab 531 \tab 979 \tab 209 \tab 1875 \tab 389 \tab 108 \tab 493 \tab 657 \tab 874 \tab 294 \tab 265 \tab 195 \tab 621 \tab 1245 \tab 632 \tab 954 \tab 208 \tab 1913 \tab 401 \tab 149 \tab 70 \tab 75 \tab 208 \tab 307 \tab 51 \tab 683 \tab 346 \tab 422 \tab 322 \tab 71 \tab 67 \tab 152 \tab 58 \tab 137 \tab 29 \tab 101 \tab 176 \tab 53 \tab 1998 \tab 7 \tab 13 \tab 1 \tab 14 \tab 11 \tab 49434 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{evwork}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{id}},
\code{\link{occ}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{year}}

Other Occupation: 
\code{\link{occ}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}}
}
\concept{Core}
\concept{Occupation}
\concept{Occupational Prestige}
\keyword{variable}
