% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_15.R
\name{preteen}
\alias{preteen}
\title{Household members 6 thru 12 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
preteen
}
\details{
Question Household Size and Composition (see Appendix D:  Recodes, for information about these variables). B. NUMBER OF MEMBERS 6 TO 12 YEARS OLD:
As of 2024, PRETEEN has been recoded to only account for 0 to 2+ children ages 6 to 12 in the household. This has been retroactively applied to all previous years of PRETEEN. The original PRETEEN containing data from 1972 to 2022 has been renamed PRETEEN_7222.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0
\item \verb{[1]} 1
\item \verb{[2]} 2 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/99/vshow}{GSS Data Explorer page for \code{preteen}}.

Counts by year:\tabular{llllllll}{
   year \tab 0 \tab 1 \tab 2 or more \tab no answer \tab skipped on web \tab iap \tab Total \cr
   1972 \tab 1159 \tab 252 \tab 202 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1101 \tab 212 \tab 191 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1066 \tab 224 \tab 192 \tab 2 \tab - \tab - \tab 1484 \cr
   1975 \tab 1127 \tab 202 \tab 161 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1139 \tab 221 \tab 137 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab 1154 \tab 219 \tab 148 \tab 9 \tab - \tab - \tab 1530 \cr
   1978 \tab 1159 \tab 220 \tab 139 \tab 14 \tab - \tab - \tab 1532 \cr
   1980 \tab 1131 \tab 209 \tab 116 \tab 12 \tab - \tab - \tab 1468 \cr
   1982 \tab 1470 \tab 259 \tab 126 \tab 5 \tab - \tab - \tab 1860 \cr
   1983 \tab 1235 \tab 228 \tab 130 \tab 6 \tab - \tab - \tab 1599 \cr
   1984 \tab 1188 \tab 169 \tab 111 \tab 5 \tab - \tab - \tab 1473 \cr
   1985 \tab 1238 \tab 184 \tab 107 \tab 5 \tab - \tab - \tab 1534 \cr
   1986 \tab 1172 \tab 184 \tab 108 \tab 6 \tab - \tab - \tab 1470 \cr
   1987 \tab 1438 \tab 237 \tab 142 \tab 2 \tab - \tab - \tab 1819 \cr
   1988 \tab 1167 \tab 196 \tab 106 \tab 12 \tab - \tab - \tab 1481 \cr
   1989 \tab 1208 \tab 193 \tab 116 \tab 20 \tab - \tab - \tab 1537 \cr
   1990 \tab 1102 \tab 157 \tab 103 \tab 10 \tab - \tab - \tab 1372 \cr
   1991 \tab 1198 \tab 184 \tab 106 \tab 29 \tab - \tab - \tab 1517 \cr
   1993 \tab 1294 \tab 192 \tab 116 \tab 4 \tab - \tab - \tab 1606 \cr
   1994 \tab 2423 \tab 369 \tab 197 \tab 3 \tab - \tab - \tab 2992 \cr
   1996 \tab 2368 \tab 346 \tab 190 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2315 \tab 311 \tab 200 \tab 6 \tab - \tab - \tab 2832 \cr
   2000 \tab 2200 \tab 377 \tab 193 \tab 47 \tab - \tab - \tab 2817 \cr
   2002 \tab 2337 \tab 260 \tab 138 \tab 30 \tab - \tab - \tab 2765 \cr
   2004 \tab 2358 \tab 282 \tab 153 \tab 19 \tab - \tab - \tab 2812 \cr
   2006 \tab 3748 \tab 483 \tab 242 \tab 37 \tab - \tab - \tab 4510 \cr
   2008 \tab 1673 \tab 219 \tab 102 \tab 29 \tab - \tab - \tab 2023 \cr
   2010 \tab 1739 \tab 186 \tab 112 \tab 7 \tab - \tab - \tab 2044 \cr
   2012 \tab 1636 \tab 210 \tab 107 \tab 21 \tab - \tab - \tab 1974 \cr
   2014 \tab 2220 \tab 195 \tab 119 \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab 2497 \tab 229 \tab 115 \tab 26 \tab - \tab - \tab 2867 \cr
   2018 \tab 2047 \tab 183 \tab 89 \tab 29 \tab - \tab - \tab 2348 \cr
   2021 \tab 3544 \tab 313 \tab 172 \tab - \tab 3 \tab - \tab 4032 \cr
   2022 \tab 1544 \tab 138 \tab 86 \tab 8 \tab 4 \tab 1764 \tab 3544 \cr
   2024 \tab 1437 \tab 159 \tab 80 \tab 1 \tab 2 \tab 1630 \tab 3309 \cr
   Total \tab 58832 \tab 8202 \tab 4852 \tab 410 \tab 9 \tab 3394 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{granborn}},
\code{\link{hompop}},
\code{\link{income}},
\code{\link{rincome}},
\code{\link{teens}},
\code{\link{unrelat}}

Other Household: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{hompop}},
\code{\link{teens}},
\code{\link{unrelat}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
