% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_233.R
\name{prfmatt}
\alias{prfmatt}
\title{Did r attend alone or with others}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmatt
}
\details{
Question Did you attend this performance alone, with one other person, or with two or more other people?
}
\section{Values}{

\itemize{
\item \verb{[1]} alone
\item \verb{[2]} with one other person
\item \verb{[3]} with two or more other people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6214/vshow}{GSS Data Explorer page for \code{prfmatt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab alone \tab with one other person \tab with two or more other people \tab no answer \tab Total \cr
   2016 \tab 2433 \tab 19 \tab 190 \tab 217 \tab 8 \tab 2867 \cr
   Total \tab 2433 \tab 19 \tab 190 \tab 217 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artexbt}},
\code{\link{prfmatt1}},
\code{\link{prfmatt2}},
\code{\link{prfmatt3}},
\code{\link{prfmatt4}},
\code{\link{prfmatt5}},
\code{\link{prfmdan}},
\code{\link{prfmmus}},
\code{\link{prfmthe}}
}
\concept{Arts and Culture}
\keyword{variable}
