% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_233.R
\name{prfmatt4}
\alias{prfmatt4}
\title{Attended with relative}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmatt4
}
\details{
Question Relative
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6218/vshow}{GSS Data Explorer page for \code{prfmatt4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 2452 \tab 125 \tab 282 \tab 8 \tab 2867 \cr
   Total \tab 2452 \tab 125 \tab 282 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artexbt}},
\code{\link{prfmatt}},
\code{\link{prfmatt1}},
\code{\link{prfmatt2}},
\code{\link{prfmatt3}},
\code{\link{prfmatt5}},
\code{\link{prfmdan}},
\code{\link{prfmmus}},
\code{\link{prfmthe}}
}
\concept{Arts and Culture}
\keyword{variable}
