% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_236.R
\name{prfmcost}
\alias{prfmcost}
\title{Did not go due to cost}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmcost
}
\details{
Question Thinking about the most recent performance you wanted to attend but did not, which of the following factors were important in your decision not to attend?     
Cost too much.
}
\section{Values}{

\itemize{
\item \code{1} a reason
\item \code{2} not a reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6249/vshow}{GSS Data Explorer page for \code{prfmcost}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{artwhy3}},
\code{\link{artwhy4}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo}},
\code{\link{nogo1}}
}
\concept{Arts and Culture}
\keyword{variable}
