% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_233.R
\name{prfmdan}
\alias{prfmdan}
\title{Was it a dance performance}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmdan
}
\details{
Question Dance
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6212/vshow}{GSS Data Explorer page for \code{prfmdan}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 2433 \tab 74 \tab 351 \tab 9 \tab 2867 \cr
   Total \tab 2433 \tab 74 \tab 351 \tab 9 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artexbt}},
\code{\link{prfmatt}},
\code{\link{prfmatt1}},
\code{\link{prfmatt2}},
\code{\link{prfmatt3}},
\code{\link{prfmatt4}},
\code{\link{prfmatt5}},
\code{\link{prfmmus}},
\code{\link{prfmthe}}
}
\concept{Arts and Culture}
\keyword{variable}
