% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_234.R
\name{prfmfree}
\alias{prfmfree}
\title{Was performance attended free}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmfree
}
\details{
Question Was the performance you attended free, or was there an admission fee?
}
\section{Values}{

\itemize{
\item \verb{[1]} free
\item \verb{[2]} admission fee
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6220/vshow}{GSS Data Explorer page for \code{prfmfree}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab free \tab admission fee \tab no answer \tab Total \cr
   2016 \tab 2433 \tab 82 \tab 344 \tab 8 \tab 2867 \cr
   Total \tab 2433 \tab 82 \tab 344 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{prfmwhy1}},
\code{\link{prfmwhy2}},
\code{\link{prfmwhy3}},
\code{\link{prfmwhy4}},
\code{\link{prfmwhy5}},
\code{\link{prfmwhy6}},
\code{\link{prfmwhy7}},
\code{\link{prfmwhy8}},
\code{\link{prfmwhy9}}
}
\concept{Arts and Culture}
\keyword{variable}
