% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_237.R
\name{prfmgst}
\alias{prfmgst}
\title{Did not go because couldn't find anyone to go with}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmgst
}
\details{
Question Could not find anyone to go with.
}
\section{Values}{

\itemize{
\item \verb{[1]} a reason
\item \verb{[2]} not a reason
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6252/vshow}{GSS Data Explorer page for \code{prfmgst}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab a reason \tab not a reason \tab don't know \tab not available in this year \tab Total \cr
   2016 \tab 2408 \tab 91 \tab 367 \tab 1 \tab - \tab 2867 \cr
   Total \tab 2408 \tab 91 \tab 367 \tab 1 \tab 0 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt}},
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt4}},
\code{\link{artatt5}},
\code{\link{artcost}},
\code{\link{artexbt}},
\code{\link{artfree}},
\code{\link{artgst}},
\code{\link{artint}},
\code{\link{artloc}},
\code{\link{artmostoth}},
\code{\link{artothr}},
\code{\link{arttime}},
\code{\link{arttrvl}},
\code{\link{artwhy1}},
\code{\link{artwhy2}},
\code{\link{artwhy3}},
\code{\link{artwhy4}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo}},
\code{\link{nogo1}},
\code{\link{prfmatt}},
\code{\link{prfmatt1}},
\code{\link{prfmatt2}},
\code{\link{prfmatt3}},
\code{\link{prfmatt4}},
\code{\link{prfmatt5}},
\code{\link{prfmcost}},
\code{\link{prfmdan}},
\code{\link{prfmfree}},
\code{\link{prfmint}},
\code{\link{prfmloc}},
\code{\link{prfmmostoth}},
\code{\link{prfmmus}},
\code{\link{prfmnce}},
\code{\link{prfmothr}},
\code{\link{prfmthe}},
\code{\link{prfmtime}},
\code{\link{prfmtrvl}},
\code{\link{prfmwhy0}},
\code{\link{prfmwhy1}},
\code{\link{prfmwhy2}},
\code{\link{prfmwhy3}},
\code{\link{prfmwhy4}},
\code{\link{prfmwhy5}},
\code{\link{prfmwhy6}},
\code{\link{prfmwhy7}},
\code{\link{prfmwhy8}},
\code{\link{prfmwhy9}}
}
\concept{Arts and Culture}
\keyword{variable}
