% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_237.R
\name{prfmint}
\alias{prfmint}
\title{Did not go since not of interest}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmint
}
\details{
Question Programs or events were not of interest.
}
\section{Values}{

\itemize{
\item \verb{[1]} a reason
\item \verb{[2]} not a reason
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6250/vshow}{GSS Data Explorer page for \code{prfmint}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab a reason \tab not a reason \tab Total \cr
   2016 \tab 2408 \tab 31 \tab 428 \tab 2867 \cr
   Total \tab 2408 \tab 31 \tab 428 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artcost}},
\code{\link{artint}},
\code{\link{arttrvl}},
\code{\link{prfmgst}},
\code{\link{prfmloc}},
\code{\link{prfmothr}},
\code{\link{prfmtime}},
\code{\link{prfmtrvl}}
}
\concept{Arts and Culture}
\keyword{variable}
