% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{prfmnce}
\alias{prfmnce}
\title{Did r go to a performance in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmnce
}
\details{
Question PAY SPECIAL ATTENTION TO THE HANDCARD NUMBER INDICATED FOR THE QUESTIONS ASKED.  DEPENDING ON THE RESPONDENT'S ANSWERS, YOU WILL NEED TO REFER HIM/HER TO HANDCARD H1, H2, H3 OR H4.    
With the exception of elementary or high school performances, did you go to a live music, theater, or dance performance, during the last 12 months?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6209/vshow}{GSS Data Explorer page for \code{prfmnce}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 605 \tab 864 \tab 8 \tab 2867 \cr
   Total \tab 1390 \tab 605 \tab 864 \tab 8 \tab 2867 \cr
}
}

\concept{Arts and Culture}
\keyword{variable}
