% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_235.R
\name{prfmwhy0}
\alias{prfmwhy0}
\title{Other major reason in decision to attend perf}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmwhy0
}
\details{
Question Please specify (immediately after PRFMWHY9)
}
\section{Values}{

\itemize{
\item \code{1} gift, free
\item \code{2} school related
\item \code{3} wanted to see specific show/exhibit
\item \code{4} variety/something different/new experience
\item \code{5} win meal, drinks
\item \code{6} entertainment, enjoyment, have fun
\item \code{7} religious
\item \code{8} for children, educational
\item \code{10} support family/friend
\item \code{11} celebrate birthday, anniversary
\item \code{12} night out, date
\item \code{13} meet, support artist
\item \code{14} work related
\item \code{20} miscellaneous other reasons
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6230/vshow}{GSS Data Explorer page for \code{prfmwhy0}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab school related \tab wanted to see specific show/exhibit \tab entertainment, enjoyment, have fun \tab religious \tab support family/friend \tab celebrate birthday, anniversary \tab night out, date \tab miscellaneous other reasons \tab no answer \tab Total \cr
   2016 \tab 2784 \tab 3 \tab 19 \tab 10 \tab 3 \tab 9 \tab 7 \tab 12 \tab 12 \tab 8 \tab 2867 \cr
   Total \tab 2784 \tab 3 \tab 19 \tab 10 \tab 3 \tab 9 \tab 7 \tab 12 \tab 12 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt}},
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt4}},
\code{\link{artatt5}},
\code{\link{artfree}},
\code{\link{artwhy1}},
\code{\link{artwhy2}}
}
\concept{Arts and Culture}
\keyword{variable}
