% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_234.R
\name{prfmwhy1}
\alias{prfmwhy1}
\title{Imp of low cost in decision to attend performance}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmwhy1
}
\details{
Question People attend performances for different reasons.
How big a reason was low cost in your decision to attend this performance?  Was it a major reason, a minor reason, or not a reason at all?
SAY, "THIS CARD SHOWS SOME REASONS THAT PEOPLE ATTEND PERFORMANCES."
}
\section{Values}{

\itemize{
\item \code{1} major reason
\item \code{2} minor reason
\item \code{3} not a reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6221/vshow}{GSS Data Explorer page for \code{prfmwhy1}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{prfmfree}},
\code{\link{prfmwhy2}},
\code{\link{prfmwhy3}},
\code{\link{prfmwhy4}},
\code{\link{prfmwhy5}},
\code{\link{prfmwhy6}},
\code{\link{prfmwhy7}},
\code{\link{prfmwhy8}},
\code{\link{prfmwhy9}}
}
\concept{Arts and Culture}
\keyword{variable}
