% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_234.R
\name{prfmwhy7}
\alias{prfmwhy7}
\title{Imp of location in decision to attend perf}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmwhy7
}
\details{
Question How big a reason was wanting to experience a performance at this location in your decision to attend this performance? Was it a major reason, a minor reason, or not a reason at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} major reason
\item \verb{[2]} minor reason
\item \verb{[3]} not a reason
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6227/vshow}{GSS Data Explorer page for \code{prfmwhy7}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab major reason \tab minor reason \tab not a reason \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2433 \tab 124 \tab 129 \tab 172 \tab 1 \tab 8 \tab 2867 \cr
   Total \tab 2433 \tab 124 \tab 129 \tab 172 \tab 1 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{prfmfree}},
\code{\link{prfmwhy1}},
\code{\link{prfmwhy2}},
\code{\link{prfmwhy3}},
\code{\link{prfmwhy4}},
\code{\link{prfmwhy5}},
\code{\link{prfmwhy6}},
\code{\link{prfmwhy8}},
\code{\link{prfmwhy9}}
}
\concept{Arts and Culture}
\keyword{variable}
