% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_234.R
\name{prfmwhy8}
\alias{prfmwhy8}
\title{Imp of specific individ performer in decision to attend perf}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmwhy8
}
\details{
Question How big a reason was wanting to see a specific, individual performer in your decision to attend this performance? Was it a major reason, a minor reason, or not a reason at all?
}
\section{Values}{

\itemize{
\item \code{1} major reason
\item \code{2} minor reason
\item \code{3} not a reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6228/vshow}{GSS Data Explorer page for \code{prfmwhy8}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab major reason \tab minor reason \tab not a reason \tab 9 \tab Total \cr
   2016 \tab 2433 \tab 234 \tab 61 \tab 131 \tab 8 \tab 2867 \cr
   Total \tab 2433 \tab 234 \tab 61 \tab 131 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{prfmfree}},
\code{\link{prfmwhy1}},
\code{\link{prfmwhy2}},
\code{\link{prfmwhy3}},
\code{\link{prfmwhy4}},
\code{\link{prfmwhy5}},
\code{\link{prfmwhy6}},
\code{\link{prfmwhy7}},
\code{\link{prfmwhy9}}
}
\concept{Arts and Culture}
\keyword{variable}
