% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{priagncy}
\alias{priagncy}
\title{Registered at a private employment agency}
\source{
General Social Survey https://gss.norc.org
}
\description{
priagncy
}
\details{
Question Thinking about the last 12 months, have you done any of the following in order to find a job?
B. Registered at a private employment agency?
}
\section{Values}{

\itemize{
\item \code{1} no
\item \code{2} yes, once or twice
\item \code{3} yes, more than twice
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4600/vshow}{GSS Data Explorer page for \code{priagncy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes, once or twice \tab yes, more than twice \tab no answer \tab don't know \tab Total \cr
   1998 \tab 2460 \tab 284 \tab 15 \tab 3 \tab 70 \tab - \tab 2832 \cr
   2006 \tab 4008 \tab 468 \tab 23 \tab 7 \tab 4 \tab - \tab 4510 \cr
   2016 \tab 2533 \tab 277 \tab 20 \tab 12 \tab 23 \tab 2 \tab 2867 \cr
   Total \tab 9001 \tab 1029 \tab 58 \tab 22 \tab 97 \tab 2 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
