% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{prideorg}
\alias{prideorg}
\title{I am proud working for my firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
prideorg
}
\details{
Question To what extent do you agree or disagree with each of the following statements?
B. I am proud to be working for my firm or organization.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4586/vshow}{GSS Data Explorer page for \code{prideorg}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 1967 \tab 192 \tab 417 \tab 165 \tab 32 \tab 5 \tab 9 \tab 45 \tab - \tab 2832 \cr
   2006 \tab 3492 \tab 369 \tab 473 \tab 120 \tab 37 \tab 16 \tab 1 \tab 2 \tab - \tab 4510 \cr
   2016 \tab 1927 \tab 339 \tab 428 \tab 126 \tab 29 \tab 9 \tab 2 \tab 7 \tab - \tab 2867 \cr
   Total \tab 7386 \tab 900 \tab 1318 \tab 411 \tab 98 \tab 30 \tab 12 \tab 54 \tab 0 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{adforjob}},
\code{\link{ageemp}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{bosswrks}},
\code{\link{chngewrk}},
\code{\link{chngwork}},
\code{\link{cowrkers}},
\code{\link{dailywrk}},
\code{\link{danger}},
\code{\link{decidwrk}},
\code{\link{discwk5}},
\code{\link{dk}},
\code{\link{dowell}},
\code{\link{econsup}},
\code{\link{econsup1}},
\code{\link{edcskill}},
\code{\link{educemp}},
\code{\link{employer}},
\code{\link{empself}},
\code{\link{evjob}},
\code{\link{exjobsat}},
\code{\link{expernc}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{famintjb}},
\code{\link{famorjob}},
\code{\link{famresp}},
\code{\link{findjob}},
\code{\link{findwork}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{getjob}},
\code{\link{harass5}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrsmoney}},
\code{\link{hwduties}},
\code{\link{indusoth}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{joborfam}},
\code{\link{jobsat}},
\code{\link{jobskill}},
\code{\link{leavejob}},
\code{\link{leisure}},
\code{\link{localnum}},
\code{\link{lookjob}},
\code{\link{lowpay}},
\code{\link{moonlite}},
\code{\link{mortravl}},
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{newjob}},
\code{\link{newskill}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{otherwrk}},
\code{\link{othhrs}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paysame}},
\code{\link{paytime}},
\code{\link{physwrk}},
\code{\link{placewrk}},
\code{\link{planwrk}},
\code{\link{priagncy}},
\code{\link{privgovt}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{pubagncy}},
\code{\link{rageemp}},
\code{\link{rcontact}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{replaceu}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rimpskls}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpaysame}},
\code{\link{rpromotn}},
\code{\link{rsecjob}},
\code{\link{rwrkindp}},
\code{\link{ryrsfirm}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexemp}},
\code{\link{skltrain}},
\code{\link{smallbig}},
\code{\link{spwrkgvt}},
\code{\link{stayorg3}},
\code{\link{stress}},
\code{\link{strngun}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{tempwork}},
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timeoff}},
\code{\link{timepdwk}},
\code{\link{timerelx}},
\code{\link{totalnum}},
\code{\link{unbetter}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unfamten}},
\code{\link{unhlthy}},
\code{\link{unionsbd}},
\code{\link{unjobexp}},
\code{\link{unjobsec}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unpleznt}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{useskill}},
\code{\link{wantads}},
\code{\link{wantjob}},
\code{\link{wantjob1}},
\code{\link{wantjob2}},
\code{\link{whydisc5}},
\code{\link{whyjbend}},
\code{\link{wkathome}},
\code{\link{wkpersnl}},
\code{\link{wktenure}},
\code{\link{wkwkends}},
\code{\link{worecsup}},
\code{\link{worknow}},
\code{\link{workweek}},
\code{\link{wornojob}},
\code{\link{worryjob}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}},
\code{\link{wrkindp}},
\code{\link{wrkshift}},
\code{\link{xhaustn}},
\code{\link{yrjobend}},
\code{\link{yrsfirm}}

Other Work: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{balwkfam}},
\code{\link{buyvalue}},
\code{\link{chngwork}},
\code{\link{cmprgmng}},
\code{\link{commute}},
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{denyrais}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{empinput}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{extr2005}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{getahead}},
\code{\link{getaheay}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{impjob}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jobcntrl}},
\code{\link{jobfind}},
\code{\link{jobfind1}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobpromo}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobskill}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{lottosay}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{monitred}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{numorg}},
\code{\link{offsup}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{parleave}},
\code{\link{pawrkslf}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{phyeffrt}},
\code{\link{placewrk}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{richwork}},
\code{\link{rimpskls}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{satjob}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexhar}},
\code{\link{skltrain}},
\code{\link{slfmangd}},
\code{\link{slsmnshp}},
\code{\link{spdays}},
\code{\link{sphedslf}},
\code{\link{sphour}},
\code{\link{spvtrfair}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{spwrkslf}},
\code{\link{stayorg3}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{talkteam}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{toldwork}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{trynewjb}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{useskill}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkcontct}},
\code{\link{wkfambiz}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkracism}},
\code{\link{wkrspns}},
\code{\link{wksexism}},
\code{\link{wksub}},
\code{\link{wksubs}},
\code{\link{wksup}},
\code{\link{wksups}},
\code{\link{wktenure}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{workdy}},
\code{\link{workfor}},
\code{\link{workhr}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{worknow}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkhome}},
\code{\link{wrkimp}},
\code{\link{wrkindep}},
\code{\link{wrkindp}},
\code{\link{wrkmuch}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{wrkwell}},
\code{\link{yearsjob}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
