% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{privacy}
\alias{privacy}
\title{Is r concerned about privacy?}
\source{
General Social Survey https://gss.norc.org
}
\description{
privacy
}
\details{
Question How concerned are you about threats to your personal privacy in America today? Would you say you are very concerned, somewhat concerned, only a little concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very concerned
\item \verb{[2]} somewhat
\item \verb{[3]} only a little
\item \verb{[4]} not concerned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/772/vshow}{GSS Data Explorer page for \code{privacy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very concerned \tab somewhat \tab only a little \tab not concerned \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 909 \tab 500 \tab 244 \tab 177 \tab 26 \tab 4 \tab 1860 \cr
   Total \tab 0 \tab 909 \tab 500 \tab 244 \tab 177 \tab 26 \tab 4 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\keyword{variable}
