% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{privgovt}
\alias{privgovt}
\title{R preference re: private public sector}
\source{
General Social Survey https://gss.norc.org
}
\description{
privgovt
}
\details{
Question Suppose you were working and could choose between different kinds of jobs. Which of the following would you personally choose?  I would advise...
Please CIRCLE ONE CODE FOR EACH.
D. Working in a private business (or) working for the government or civil service
}
\section{Values}{

\itemize{
\item \verb{[1]} private firm
\item \verb{[2]} govt
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4527/vshow}{GSS Data Explorer page for \code{privgovt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab private firm \tab govt \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 821 \tab 394 \tab 201 \tab 37 \tab 1537 \cr
   1998 \tab 1609 \tab 801 \tab 231 \tab 161 \tab 30 \tab 2832 \cr
   2006 \tab 2992 \tab 1009 \tab 449 \tab 55 \tab 5 \tab 4510 \cr
   Total \tab 4685 \tab 2631 \tab 1074 \tab 417 \tab 72 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
