% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{prnt}
\alias{prnt}
\title{R is a parent in some capacity in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
prnt
}
\details{
Question Calculated variable: The respondent answered yes to any of RBIOPRNT, RADPTPRNT, and RSTPPRNT
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8017/vshow}{GSS Data Explorer page for \code{prnt}}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab yes \tab no \tab iap \tab skipped on web \tab Total \cr
   2022 \tab - \tab 432 \tab 94 \tab 3016 \tab 2 \tab 3544 \cr
   2024 \tab - \tab 402 \tab - \tab 2907 \tab - \tab 3309 \cr
   Total \tab 0 \tab 834 \tab 94 \tab 5923 \tab 2 \tab 6853 \cr
}
}

\keyword{variable}
