% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{prntbhav}
\alias{prntbhav}
\title{Mother vs father teach child to behave}
\source{
General Social Survey https://gss.norc.org
}
\description{
prntbhav
}
\details{
Question (In your opinion, which parent should do the following? Who should...) ...teach the child how to behave?
}
\section{Values}{

\itemize{
\item \verb{[1]} the mother mostly
\item \verb{[2]} the mother somewhat more than the father
\item \verb{[3]} the mother and father equally
\item \verb{[4]} the father somewhat more than the mother
\item \verb{[5]} the father mostly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7771/vshow}{GSS Data Explorer page for \code{prntbhav}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab the mother mostly \tab the mother somewhat more than the father \tab the mother and father equally \tab the father somewhat more than the mother \tab the father mostly \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2383 \tab 9 \tab 15 \tab 1053 \tab 18 \tab 7 \tab 48 \tab 4 \tab 7 \tab 3544 \cr
   Total \tab 2383 \tab 9 \tab 15 \tab 1053 \tab 18 \tab 7 \tab 48 \tab 4 \tab 7 \tab 3544 \cr
}
}

\keyword{variable}
