% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{prntlk}
\alias{prntlk}
\title{Mother vs father care for child}
\source{
General Social Survey https://gss.norc.org
}
\description{
prntlk
}
\details{
Question People have different opinions about how well mothers and fathers are suited to look after their children. Which of the following statements comes closest to your opinion?
}
\section{Values}{

\itemize{
\item \code{1} mothers are much better suited
\item \code{2} mothers are somewhat better suited
\item \code{3} mothers and fathers are equally suited
\item \code{4} fathers are somewhat better suited
\item \code{5} fathers are much better suited
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7767/vshow}{GSS Data Explorer page for \code{prntlk}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab mothers are much better suited \tab mothers are somewhat better suited \tab mothers and fathers are equally suited \tab fathers are somewhat better suited \tab fathers are much better suited \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2380 \tab 98 \tab 178 \tab 794 \tab 6 \tab 5 \tab 69 \tab 5 \tab 9 \tab 3544 \cr
   Total \tab 2380 \tab 98 \tab 178 \tab 794 \tab 6 \tab 5 \tab 69 \tab 5 \tab 9 \tab 3544 \cr
}
}

\keyword{variable}
