% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_560.R
\name{prob1}
\alias{prob1}
\title{First problem mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
prob1
}
\details{
Question What are the most important problems that you and members of your household have had during the last 12 months?
RECORD VERBATIM.
PROBE FOR ADDITIONAL MENTIONS AND FOR SPECIFIC DETAILS. TRY NOT TO ACCEPT GENERAL RESPONSES LIKE "WORK" OR "CRIME."  PROBE TO FIND WHAT ABOUT "WORK" IS THE PROBLEM. FOR EXAMPLE, IS IT JOB INSECURITY, LOW PAY, ARGUMENTS WITH THE BOSS, OR WHAT? OR ABOUT "CRIME", WAS THERE A VICTIMIZATION?  WHO WAS VICTIMIZED AND IN WHAT WAY?
PROBE FOR ADDITIONAL PROBLEMS.
A. First problem mentioned.
}
\section{Values}{

\itemize{
\item \verb{[995]} cant understand answer
\item \verb{[996]} no problems
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 1991, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5132/vshow}{GSS Data Explorer page for \code{prob1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 103 \tab 104 \tab 11 \tab 110 \tab 120 \tab 130 \tab 131 \tab 150 \tab 151 \tab 170 \tab 172 \tab 175 \tab 180 \tab 190 \tab 192 \tab 2 \tab 20 \tab 200 \tab 21 \tab 210 \tab 211 \tab 220 \tab 23 \tab 235 \tab 237 \tab 24 \tab 240 \tab 25 \tab 26 \tab 27 \tab 28 \tab 30 \tab 4 \tab 40 \tab 44 \tab 61 \tab 65 \tab 66 \tab 7 \tab 70 \tab 72 \tab 73 \tab 74 \tab 76 \tab 77 \tab 80 \tab 81 \tab 83 \tab 84 \tab 85 \tab 90 \tab 92 \tab cant understand answer \tab no problems \tab no answer \tab 100 \tab 101 \tab 115 \tab 132 \tab 133 \tab 140 \tab 191 \tab 201 \tab 29 \tab 3 \tab 31 \tab 43 \tab 47 \tab 6 \tab 86 \tab 94 \tab 96 \tab don't know \tab not available in this year \tab Total \cr
   1991 \tab 1025 \tab 39 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 4 \tab 1 \tab 3 \tab 3 \tab 3 \tab 6 \tab 3 \tab 2 \tab 1 \tab 43 \tab 11 \tab 7 \tab 1 \tab 6 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 9 \tab 1 \tab 2 \tab 99 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 7 \tab 1 \tab 16 \tab 3 \tab 3 \tab 1 \tab 7 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 113 \tab 42 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   2004 \tab 2378 \tab 40 \tab 15 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 7 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 5 \tab 3 \tab 26 \tab 13 \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab 18 \tab - \tab 2 \tab 88 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab 9 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 4 \tab - \tab 100 \tab 9 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 4 \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 2812 \cr
   Total \tab 3403 \tab 79 \tab 20 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab 9 \tab 5 \tab 2 \tab 3 \tab 5 \tab 6 \tab 9 \tab 4 \tab 7 \tab 4 \tab 69 \tab 24 \tab 9 \tab 5 \tab 9 \tab 2 \tab 1 \tab 6 \tab 4 \tab 3 \tab 1 \tab 8 \tab 3 \tab 27 \tab 1 \tab 4 \tab 187 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 12 \tab 3 \tab 25 \tab 4 \tab 5 \tab 1 \tab 8 \tab 3 \tab 6 \tab 2 \tab 1 \tab 3 \tab 6 \tab 1 \tab 213 \tab 51 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 4 \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab 0 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{extra1}},
\code{\link{extra2}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famper6}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth12}},
\code{\link{hlth13}},
\code{\link{hlth14}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}},
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}},
\code{\link{oth5}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{work1}},
\code{\link{work10}},
\code{\link{work2}},
\code{\link{work3}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}

Other Life Events: 
\code{\link{attrelig}},
\code{\link{eatout}},
\code{\link{extra1}},
\code{\link{extra2}},
\code{\link{famper6}},
\code{\link{godoc}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{seefilm}},
\code{\link{solveown}}

Other Trauma: 
\code{\link{clsclsex}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp4}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}},
\code{\link{law1}},
\code{\link{law2}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}},
\code{\link{law6}},
\code{\link{law7}},
\code{\link{live1}},
\code{\link{live2}},
\code{\link{live3}},
\code{\link{live4}},
\code{\link{oth1}},
\code{\link{oth2}},
\code{\link{oth3}},
\code{\link{oth4}},
\code{\link{oth5}},
\code{\link{outofbiz}},
\code{\link{ownclerg}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unemp5}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{work1}},
\code{\link{work10}},
\code{\link{work2}},
\code{\link{work3}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}
}
\concept{Life Events}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
