% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_209.R
\name{probfix2}
\alias{probfix2}
\title{Second mention of problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
probfix2
}
\details{
Question Problems often come up in life. Sometimes they're personal problems -- people are very unhappy, or nervous and irritable all the time. Sometimes there are problems in a marriage -- a husband and wife just can't get along with each other. Or, sometimes it's a personal problem with a child or a job. I'd like to ask you a few questions now about what you think a person might do to handle problems like this. For instance, let's suppose you had a lot of personal problems and you're very unhappy all the time. Let's suppose you've been that way for a long time, and it isn't getting any better. What do you think you'd Do about it?
2. What R would do about the problem -- Second mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1981/vshow}{GSS Data Explorer page for \code{probfix2}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab 11 \tab 21 \tab 22 \tab 26 \tab 31 \tab 41 \tab 51 \tab 61 \tab 62 \tab 63 \tab 64 \tab 68 \tab 69 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 81 \tab 82 \tab don't know \tab no answer \tab 23 \tab 65 \tab 95 \tab Total \cr
   1996 \tab 2448 \tab 1 \tab 1 \tab 9 \tab 1 \tab 7 \tab 1 \tab 23 \tab 6 \tab 3 \tab 1 \tab 7 \tab 2 \tab 12 \tab 18 \tab 54 \tab 7 \tab 13 \tab 6 \tab 119 \tab 136 \tab 2 \tab 27 \tab - \tab - \tab - \tab 2904 \cr
   2006 \tab 3776 \tab - \tab - \tab 8 \tab - \tab 1 \tab - \tab 12 \tab 7 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 3 \tab 17 \tab 4 \tab 3 \tab 3 \tab 53 \tab 65 \tab - \tab 39 \tab 1 \tab 2 \tab 508 \tab 4510 \cr
   Total \tab 6224 \tab 1 \tab 1 \tab 17 \tab 1 \tab 8 \tab 1 \tab 35 \tab 13 \tab 4 \tab 1 \tab 8 \tab 5 \tab 15 \tab 21 \tab 71 \tab 11 \tab 16 \tab 9 \tab 172 \tab 201 \tab 2 \tab 66 \tab 1 \tab 2 \tab 508 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{probfix1}}
}
\concept{Mental Health}
\keyword{variable}
