% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{probfix3}
\alias{probfix3}
\title{Third mention of problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
probfix3
}
\details{
Question Problems often come up in life. Sometimes they're personal problems -- people are very unhappy, or nervous and irritable all the time. Sometimes there are problems in a marriage -- a husband and wife just can't get along with each other. Or, sometimes it's a personal problem with a child or a job. I'd like to ask you a few questions now about what you think a person might do to handle problems like this. For instance, let's suppose you had a lot of personal problems and you're very unhappy all the time. Let's suppose you've been that way for a long time, and it isn't getting any better. What do you think you'd Do about it?
3. What R would do about the problem -- Third mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1982/vshow}{GSS Data Explorer page for \code{probfix3}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 22 \tab 31 \tab 51 \tab 61 \tab 62 \tab 64 \tab 65 \tab 67 \tab 68 \tab 69 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 81 \tab 82 \tab don't know \tab no answer \tab 95 \tab Total \cr
   1996 \tab 2684 \tab 10 \tab 1 \tab 14 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 3 \tab 7 \tab 17 \tab 3 \tab 5 \tab 3 \tab 61 \tab 55 \tab 2 \tab 28 \tab - \tab 2904 \cr
   2006 \tab 3776 \tab 3 \tab - \tab 6 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 22 \tab 17 \tab - \tab 39 \tab 632 \tab 4510 \cr
   Total \tab 6460 \tab 13 \tab 1 \tab 20 \tab 3 \tab 1 \tab 7 \tab 3 \tab 1 \tab 3 \tab 5 \tab 8 \tab 19 \tab 5 \tab 6 \tab 3 \tab 83 \tab 72 \tab 2 \tab 67 \tab 632 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
