% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{probhlp1}
\alias{probhlp1}
\title{First mention to help problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
probhlp1
}
\details{
Question Problems often come up in life. Sometimes they're personal problems -- people are very unhappy, or nervous and irritable all the time. Sometimes there are problems in a marriage -- a husband and wife just can't get along with each other. Or, sometimes it's a personal problem with a child or a job. I'd like to ask you a few questions now about what you think a person might do to handle problems like this. For instance, let's suppose you had a lot of personal problems and you're very unhappy all the time. Let's suppose you've been that way for a long time, and it isn't getting any better. What do you think you'd Do about it?
\enumerate{
\item The sources of help R would seek -- First mention.
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1983/vshow}{GSS Data Explorer page for \code{probhlp1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 100 \tab 103 \tab 104 \tab 105 \tab 106 \tab 107 \tab 108 \tab 190 \tab 200 \tab 201 \tab 202 \tab 290 \tab 300 \tab 390 \tab 400 \tab 401 \tab 402 \tab 403 \tab 490 \tab 499 \tab 500 \tab 502 \tab 503 \tab 505 \tab 506 \tab 509 \tab 590 \tab 602 \tab 608 \tab 690 \tab 700 \tab 800 \tab 980 \tab no answer \tab 102 \tab 910 \tab not available in this year \tab Total \cr
   1996 \tab 2168 \tab 16 \tab 8 \tab 1 \tab 2 \tab 27 \tab 9 \tab 2 \tab 5 \tab 51 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab 41 \tab 3 \tab 2 \tab 27 \tab 14 \tab 27 \tab 25 \tab 8 \tab 1 \tab 3 \tab 158 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 6 \tab 1 \tab 274 \tab - \tab - \tab - \tab 2904 \cr
   2006 \tab 3776 \tab 11 \tab 3 \tab - \tab 1 \tab 27 \tab - \tab - \tab - \tab 23 \tab 1 \tab - \tab - \tab - \tab 1 \tab 19 \tab 2 \tab - \tab 96 \tab 40 \tab - \tab 15 \tab 7 \tab - \tab - \tab 111 \tab 11 \tab - \tab - \tab - \tab 2 \tab 10 \tab - \tab 88 \tab 40 \tab 1 \tab 225 \tab - \tab 4510 \cr
   Total \tab 5944 \tab 27 \tab 11 \tab 1 \tab 3 \tab 54 \tab 9 \tab 2 \tab 5 \tab 74 \tab 11 \tab 1 \tab 2 \tab 1 \tab 3 \tab 60 \tab 5 \tab 2 \tab 123 \tab 54 \tab 27 \tab 40 \tab 15 \tab 1 \tab 3 \tab 269 \tab 12 \tab 1 \tab 1 \tab 1 \tab 5 \tab 12 \tab 6 \tab 89 \tab 314 \tab 1 \tab 225 \tab 0 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{avoidmh}},
\code{\link{breakdwn}},
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{brkddo1}},
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{brkdtime}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{charactr}},
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_oth}},
\code{\link{chldshop_wht}},
\code{\link{chldvig}},
\code{\link{cutself}},
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{depress}},
\code{\link{diagnosd}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{doclist}},
\code{\link{docsui}},
\code{\link{dofirst}},
\code{\link{evbrkdwn}},
\code{\link{evmhp}},
\code{\link{fammhneg}},
\code{\link{fekids5}},
\code{\link{fighpeer}},
\code{\link{foodalc}},
\code{\link{genetics}},
\code{\link{gesttalk}},
\code{\link{godswill}},
\code{\link{govmentl}},
\code{\link{hlthdays}},
\code{\link{hlthinsr}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hospice}},
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imbalnce}},
\code{\link{imprvown}},
\code{\link{imprvown1}},
\code{\link{imprvtrt}},
\code{\link{instrctn}},
\code{\link{knwmhosp}},
\code{\link{knwnproz}},
\code{\link{knwpatnt}},
\code{\link{medaddct}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{meddoc}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentlill}},
\code{\link{mentloth}},
\code{\link{mhaccptd}},
\code{\link{mhbrain}},
\code{\link{mhcause}},
\code{\link{mhchrctr}},
\code{\link{mhclsoth}},
\code{\link{mhcreatv}},
\code{\link{mhdecslf}},
\code{\link{mhdoactv}},
\code{\link{mhdodiet}},
\code{\link{mhdodoc}},
\code{\link{mhdoexrc}},
\code{\link{mhdofam}},
\code{\link{mhdofrgt}},
\code{\link{mhdofrnd}},
\code{\link{mhdogrp}},
\code{\link{mhdoheal}},
\code{\link{mhdoherb}},
\code{\link{mhdohosp}},
\code{\link{mhdolife}},
\code{\link{mhdomhp}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdootc}},
\code{\link{mhdooth}},
\code{\link{mhdopray}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhdorx}},
\code{\link{mhexpoth}},
\code{\link{mhfamshm}},
\code{\link{mhfeelem}},
\code{\link{mhfriend}},
\code{\link{mhfright}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhgvtjob}},
\code{\link{mhhired}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlpoth}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhhrdtlk}},
\code{\link{mhhsptrt}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhkdcare}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhluck}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhmoney}},
\code{\link{mhneihbr}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhnotch}},
\code{\link{mhnotell}},
\code{\link{mhnotpol}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhoutsdr}},
\code{\link{mhphyscl}},
\code{\link{mhprdctv}},
\code{\link{mhproblm}},
\code{\link{mhraised}},
\code{\link{mhreloth}},
\code{\link{mhresoth}},
\code{\link{mhsecret}},
\code{\link{mhseedoc}},
\code{\link{mhseepub}},
\code{\link{mhseroth}},
\code{\link{mhslfshm}},
\code{\link{mhsocial}},
\code{\link{mhstress}},
\code{\link{mhsupwrk}},
\code{\link{mhsymp}},
\code{\link{mhtrtot2}},
\code{\link{mhtrtoth}},
\code{\link{mhtrtslf}},
\code{\link{mhtrusty}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}},
\code{\link{mhupdown}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}},
\code{\link{misswork}},
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}},
\code{\link{mnilwhy3}},
\code{\link{mntlas1}},
\code{\link{mntlas2}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{mntlhlth}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlrel}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}},
\code{\link{morextrm}},
\code{\link{mustdoc}},
\code{\link{mustdoc1}},
\code{\link{musthosp}},
\code{\link{musthosp1}},
\code{\link{mustmed}},
\code{\link{mustmed1}},
\code{\link{negself}},
\code{\link{notreat}},
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}},
\code{\link{otcmed}},
\code{\link{othmhneg}},
\code{\link{painmed}},
\code{\link{physill}},
\code{\link{physill1}},
\code{\link{probfix1}},
\code{\link{probfix2}},
\code{\link{probfix3}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozac}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozkid3}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}},
\code{\link{pryntfam}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psyckid3}},
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{psycpsyc}},
\code{\link{putsoff}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}},
\code{\link{rightref}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{rxmed}},
\code{\link{schshoot}},
\code{\link{seemhpub}},
\code{\link{seenmntl}},
\code{\link{selfhelp}},
\code{\link{seriousp}},
\code{\link{smedia}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sphealer}},
\code{\link{spmentl}},
\code{\link{stresses}},
\code{\link{swithlth}},
\code{\link{takelife}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}},
\code{\link{termvig}},
\code{\link{tlkclrgy}},
\code{\link{tlkfam}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}},
\code{\link{upsdowns}},
\code{\link{usedmntl}},
\code{\link{usedproz}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{viglab06}},
\code{\link{viglabel}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigvermy}},
\code{\link{vigversn}},
\code{\link{vigwork}},
\code{\link{wayraise}}

Other Mental Health: 
\code{\link{anydoc}},
\code{\link{avoidmh}},
\code{\link{breakdwn}},
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{brkddo1}},
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{brkdtime}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{charactr}},
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_oth}},
\code{\link{chldshop_wht}},
\code{\link{chldvig}},
\code{\link{cutself}},
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{depress}},
\code{\link{diagnosd}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{doclist}},
\code{\link{docsui}},
\code{\link{dofirst}},
\code{\link{evbrkdwn}},
\code{\link{evmhp}},
\code{\link{fammhneg}},
\code{\link{fekids5}},
\code{\link{fighpeer}},
\code{\link{foodalc}},
\code{\link{genetics}},
\code{\link{gesttalk}},
\code{\link{godswill}},
\code{\link{govmentl}},
\code{\link{hlthdays}},
\code{\link{hlthinsr}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hospice}},
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imbalnce}},
\code{\link{imprvown}},
\code{\link{imprvown1}},
\code{\link{imprvtrt}},
\code{\link{instrctn}},
\code{\link{knwmhosp}},
\code{\link{knwnproz}},
\code{\link{knwpatnt}},
\code{\link{medaddct}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{meddoc}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentlill}},
\code{\link{mentloth}},
\code{\link{mhaccptd}},
\code{\link{mhbrain}},
\code{\link{mhcause}},
\code{\link{mhchrctr}},
\code{\link{mhclsoth}},
\code{\link{mhcreatv}},
\code{\link{mhdecslf}},
\code{\link{mhdoactv}},
\code{\link{mhdodiet}},
\code{\link{mhdodoc}},
\code{\link{mhdoexrc}},
\code{\link{mhdofam}},
\code{\link{mhdofrgt}},
\code{\link{mhdofrnd}},
\code{\link{mhdogrp}},
\code{\link{mhdoheal}},
\code{\link{mhdoherb}},
\code{\link{mhdohosp}},
\code{\link{mhdolife}},
\code{\link{mhdomhp}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdootc}},
\code{\link{mhdooth}},
\code{\link{mhdopray}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhdorx}},
\code{\link{mhexpoth}},
\code{\link{mhfamshm}},
\code{\link{mhfeelem}},
\code{\link{mhfriend}},
\code{\link{mhfright}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhgvtjob}},
\code{\link{mhhired}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlpoth}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhhrdtlk}},
\code{\link{mhhsptrt}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhkdcare}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhluck}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhmoney}},
\code{\link{mhneihbr}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhnotch}},
\code{\link{mhnotell}},
\code{\link{mhnotpol}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhoutsdr}},
\code{\link{mhphyscl}},
\code{\link{mhprdctv}},
\code{\link{mhproblm}},
\code{\link{mhraised}},
\code{\link{mhreloth}},
\code{\link{mhresoth}},
\code{\link{mhsecret}},
\code{\link{mhseedoc}},
\code{\link{mhseepub}},
\code{\link{mhseroth}},
\code{\link{mhslfshm}},
\code{\link{mhsocial}},
\code{\link{mhstress}},
\code{\link{mhsupwrk}},
\code{\link{mhsymp}},
\code{\link{mhtrtot2}},
\code{\link{mhtrtoth}},
\code{\link{mhtrtslf}},
\code{\link{mhtrusty}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}},
\code{\link{mhupdown}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}},
\code{\link{misswork}},
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}},
\code{\link{mnilwhy3}},
\code{\link{mntlas1}},
\code{\link{mntlas2}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{mntlhlth}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlrel}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}},
\code{\link{morextrm}},
\code{\link{mustdoc}},
\code{\link{mustdoc1}},
\code{\link{musthosp}},
\code{\link{musthosp1}},
\code{\link{mustmed}},
\code{\link{mustmed1}},
\code{\link{negself}},
\code{\link{notreat}},
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}},
\code{\link{otcmed}},
\code{\link{othmhneg}},
\code{\link{painmed}},
\code{\link{physill}},
\code{\link{physill1}},
\code{\link{probfix1}},
\code{\link{probfix2}},
\code{\link{probfix3}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozac}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozkid3}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}},
\code{\link{pryntfam}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psyckid3}},
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{psycpsyc}},
\code{\link{putsoff}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}},
\code{\link{rightref}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{rxmed}},
\code{\link{schshoot}},
\code{\link{seemhpub}},
\code{\link{seenmntl}},
\code{\link{selfhelp}},
\code{\link{seriousp}},
\code{\link{smedia}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sphealer}},
\code{\link{spmentl}},
\code{\link{stresses}},
\code{\link{swithlth}},
\code{\link{takelife}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}},
\code{\link{termvig}},
\code{\link{tlkclrgy}},
\code{\link{tlkfam}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}},
\code{\link{upsdowns}},
\code{\link{usedmntl}},
\code{\link{usedproz}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{viglab06}},
\code{\link{viglabel}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigvermy}},
\code{\link{vigversn}},
\code{\link{vigwork}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
