% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{probhlp3}
\alias{probhlp3}
\title{Third mention to help problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
probhlp3
}
\details{
Question Problems often come up in life. Sometimes they're personal problems -- people are very unhappy, or nervous and irritable all the time. Sometimes there are problems in a marriage -- a husband and wife just can't get along with each other. Or, sometimes it's a personal problem with a child or a job. I'd like to ask you a few questions now about what you think a person might do to handle problems like this. For instance, let's suppose you had a lot of personal problems and you're very unhappy all the time. Let's suppose you've been that way for a long time, and it isn't getting any better. What do you think you'd Do about it?
3. The sources of help R would seek -- Third mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1985/vshow}{GSS Data Explorer page for \code{probhlp3}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab 100 \tab 101 \tab 104 \tab 106 \tab 107 \tab 108 \tab 190 \tab 200 \tab 201 \tab 290 \tab 390 \tab 400 \tab 401 \tab 403 \tab 490 \tab 499 \tab 500 \tab 502 \tab 506 \tab 605 \tab 690 \tab 700 \tab 800 \tab no answer \tab 103 \tab 105 \tab 910 \tab 980 \tab 995 \tab Total \cr
   1996 \tab 2772 \tab 7 \tab 1 \tab 1 \tab 9 \tab 1 \tab 6 \tab 4 \tab 11 \tab 3 \tab 1 \tab 2 \tab 10 \tab 2 \tab 7 \tab 2 \tab 4 \tab 4 \tab 1 \tab 9 \tab 1 \tab 2 \tab 3 \tab 3 \tab 38 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   2006 \tab 3776 \tab 5 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 10 \tab - \tab - \tab 1 \tab 4 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab 38 \tab 1 \tab 1 \tab 22 \tab 4 \tab 628 \tab 4510 \cr
   Total \tab 6548 \tab 12 \tab 2 \tab 2 \tab 12 \tab 2 \tab 6 \tab 5 \tab 21 \tab 3 \tab 1 \tab 3 \tab 14 \tab 2 \tab 10 \tab 2 \tab 4 \tab 5 \tab 3 \tab 16 \tab 1 \tab 2 \tab 3 \tab 3 \tab 76 \tab 1 \tab 1 \tab 22 \tab 4 \tab 628 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
