% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_290.R
\name{processr}
\alias{processr}
\title{R knows what a processor is}
\source{
General Social Survey https://gss.norc.org
}
\description{
processr
}
\details{
Question A. Do you know whether the "486" or "pentium" on a personal computer (and G3 on a Mac) refer to...
}
\section{Values}{

\itemize{
\item \verb{[1]} the size of the hard drive
\item \verb{[2]} the speed of the hard drive
\item \verb{[3]} the type of microprocessor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2630/vshow}{GSS Data Explorer page for \code{processr}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab the size of the hard drive \tab the speed of the hard drive \tab the type of microprocessor \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2142 \tab 63 \tab 115 \tab 256 \tab 140 \tab 49 \tab 2765 \cr
   Total \tab 2142 \tab 63 \tab 115 \tab 256 \tab 140 \tab 49 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{askelse}},
\code{\link{askwksch}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{cussup}},
\code{\link{doself}},
\code{\link{manbook}},
\code{\link{payelse}},
\code{\link{tecsup}}
}
\concept{Information Society}
\keyword{variable}
