% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_307.R
\name{prodctiv}
\alias{prodctiv}
\title{Work conditions allow productivity}
\source{
General Social Survey https://gss.norc.org
}
\description{
prodctiv
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
P. Conditions on my job allow me to be about as productive as I could be
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2789/vshow}{GSS Data Explorer page for \code{prodctiv}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{partteam}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{safehlth}},
\code{\link{setthngs}},
\code{\link{toofewwk}},
\code{\link{trdunion}},
\code{\link{wkdecide}},
\code{\link{wksmooth}}

Other Job: 
\code{\link{partteam}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{safehlth}},
\code{\link{setthngs}},
\code{\link{toofewwk}},
\code{\link{wkdecide}},
\code{\link{wksmooth}}
}
\concept{Job}
\concept{Quality of Working Life}
\keyword{variable}
