% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{profits2}
\alias{profits2}
\title{Should co. pay workers more than shareholders}
\source{
General Social Survey https://gss.norc.org
}
\description{
profits2
}
\details{
Question To what extent do you agree or disagree with the following statements?
Please tell me whether you strongly agree, agree, disagree, or strongly disagree:
B. Corporations should pay more of their profits to workers and less to shareholders…
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1264/vshow}{GSS Data Explorer page for \code{profits2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab - \tab 248 \tab 895 \tab 255 \tab 18 \tab 94 \tab 7 \tab - \tab 1517 \cr
   Total \tab 0 \tab 248 \tab 895 \tab 255 \tab 18 \tab 94 \tab 7 \tab 0 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{afailure}},
\code{\link{automatn}},
\code{\link{baseofop}},
\code{\link{chngeorg}},
\code{\link{cowrkrel}},
\code{\link{drugtest}},
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{futpromo}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{helporg}},
\code{\link{idecide}},
\code{\link{imatter}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobcntrl}},
\code{\link{jobjudge}},
\code{\link{jobvshme}},
\code{\link{jobyears}},
\code{\link{learning}},
\code{\link{levels}()},
\code{\link{lottosay}},
\code{\link{ltrsref}},
\code{\link{merged}},
\code{\link{monitred}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{offsup}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{orgfin}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{othpay}},
\code{\link{outofbiz}},
\code{\link{pessimst}},
\code{\link{physical}},
\code{\link{profits1}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{proudorg}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{racpromo}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{reorg}},
\code{\link{samejob}},
\code{\link{samevals}},
\code{\link{satself}},
\code{\link{schoolng}},
\code{\link{sexpromo}},
\code{\link{skiltest}},
\code{\link{slfrspct}},
\code{\link{splocnum}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{totunemp}},
\code{\link{training}},
\code{\link{unmanrel}},
\code{\link{unpower}},
\code{\link{unprog}},
\code{\link{unvote}},
\code{\link{wherewrk}},
\code{\link{wrkindep}},
\code{\link{wrkmuch}},
\code{\link{wrkwell}},
\code{\link{wrkyears}},
\code{\link{youmoney}},
\code{\link{yousup}}

Other Corporations: 
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{equal2}},
\code{\link{equal6}},
\code{\link{merged}},
\code{\link{orgfin}},
\code{\link{outofbiz}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{profits1}},
\code{\link{reorg}},
\code{\link{startbiz}},
\code{\link{whybiz}},
\code{\link{yearbiz}}

Other Social Class: 
\code{\link{books16}},
\code{\link{bossemps}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{conage}},
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{educop}},
\code{\link{equal1}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{goveqinc}},
\code{\link{profits1}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{wlthpov}}
}
\concept{Corporations}
\concept{Social Class}
\concept{Work Organizations}
\keyword{variable}
