% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{progtax}
\alias{progtax}
\title{How much income tax should rich pay?}
\source{
General Social Survey https://gss.norc.org
}
\description{
progtax
}
\details{
Question Some people think those with high incomes should pay a larger proportion (percentage) of their earnings in taxes than those who earn low incomes. Other people think that those with high incomes and those with low incomes should pay the same proportion (percentage) of their earnings in taxes. Do you think those with high incomes should pay a much larger proportion, pay a larger proportion, pay the same proportion as those who earn low incomes, pay a smaller proportion. pay a much smaller proportion?
}
\section{Values}{

\itemize{
\item \verb{[1]} pay much more
\item \verb{[2]} pay more
\item \verb{[3]} all pay same
\item \verb{[4]} pay less
\item \verb{[5]} pay much less
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4004/vshow}{GSS Data Explorer page for \code{progtax}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab pay much more \tab pay more \tab all pay same \tab pay less \tab pay much less \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1985 \tab 857 \tab 119 \tab 252 \tab 265 \tab 6 \tab 1 \tab 25 \tab 9 \tab - \tab 1534 \cr
   1990 \tab 155 \tab 218 \tab 497 \tab 425 \tab 10 \tab 5 \tab 51 \tab 11 \tab - \tab 1372 \cr
   Total \tab 1012 \tab 337 \tab 749 \tab 690 \tab 16 \tab 6 \tab 76 \tab 20 \tab 0 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidneedy}},
\code{\link{aidold}},
\code{\link{aidsmart}},
\code{\link{aidunemp}},
\code{\link{beltup}},
\code{\link{bribe}},
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{buspow}},
\code{\link{bustax}},
\code{\link{cctv}},
\code{\link{colop}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{cutdebt}},
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{databank}},
\code{\link{demworks}},
\code{\link{emonitor}},
\code{\link{eqincome}},
\code{\link{equalize}},
\code{\link{exploit}},
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{fewtrsty}},
\code{\link{gendereq}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{govtinfo}},
\code{\link{govtpow}},
\code{\link{grnlaws}},
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}},
\code{\link{hlphitec}},
\code{\link{hlpinflu}},
\code{\link{hlthcare}},
\code{\link{hsbasics}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{jobsall}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidout}},
\code{\link{kidskips}},
\code{\link{kidxfilm}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{mustret}},
\code{\link{nosmoke}},
\code{\link{obeylaw}},
\code{\link{opfamily}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{polgbeco}},
\code{\link{polint}},
\code{\link{polsfair}},
\code{\link{prespop}},
\code{\link{pricecon}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{protstrs}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{prvdschl}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{racists}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{ratetone}},
\code{\link{revoltrs}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}},
\code{\link{unionsok}},
\code{\link{verdict}},
\code{\link{volacty2}},
\code{\link{volactyr}},
\code{\link{wotrial}},
\code{\link{youinflu}}

Other Taxes: 
\code{\link{bustax}},
\code{\link{inctax}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{tax}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\concept{Taxes}
\keyword{variable}
