% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_634.R
\name{prolife}
\alias{prolife}
\title{I consider myself pro-life.}
\source{
General Social Survey https://gss.norc.org
}
\description{
prolife
}
\details{
Question I consider myself pro-life.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7543/vshow}{GSS Data Explorer page for \code{prolife}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 424 \tab 660 \tab 791 \tab 941 \tab 574 \tab 571 \tab 7 \tab 6 \tab 58 \tab 4032 \cr
   Total \tab 424 \tab 660 \tab 791 \tab 941 \tab 574 \tab 571 \tab 7 \tab 6 \tab 58 \tab 4032 \cr
}
}

\keyword{variable}
