% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{promotng}
\alias{promotng}
\title{Procedures for promotion}
\source{
General Social Survey https://gss.norc.org
}
\description{
promotng
}
\details{
Question Are there regular procedures for promoting people in your job to a higher level?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1299/vshow}{GSS Data Explorer page for \code{promotng}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 406 \tab 458 \tab 38 \tab 10 \tab 1517 \cr
   Total \tab 605 \tab 406 \tab 458 \tab 38 \tab 10 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobyears}},
\code{\link{samejob}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}
}
\concept{Work Organizations}
\keyword{variable}
