% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_307.R
\name{promteok}
\alias{promteok}
\title{R's chances for promotion good}
\source{
General Social Survey https://gss.norc.org
}
\description{
promteok
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
A. The chances for promotion are good
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not too true
\item \verb{[4]} not at all true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2796/vshow}{GSS Data Explorer page for \code{promteok}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 364 \tab 604 \tab 401 \tab 375 \tab 31 \tab 21 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 327 \tab 604 \tab 389 \tab 343 \tab 48 \tab 23 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 197 \tab 421 \tab 272 \tab 252 \tab 14 \tab 31 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 250 \tab 452 \tab 307 \tab 218 \tab 12 \tab 11 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 295 \tab 534 \tab 329 \tab 238 \tab 17 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 1551 \tab 401 \tab 785 \tab 431 \tab 303 \tab 38 \tab 15 \tab 20 \tab 3544 \cr
   Total \tab 8341 \tab 1834 \tab 3400 \tab 2129 \tab 1729 \tab 160 \tab 136 \tab 20 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{partteam}},
\code{\link{prodctiv}},
\code{\link{proudemp}},
\code{\link{safehlth}},
\code{\link{setthngs}},
\code{\link{toofewwk}},
\code{\link{trdunion}},
\code{\link{wkdecide}},
\code{\link{wksmooth}}
}
\concept{Quality of Working Life}
\keyword{variable}
