% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_326.R
\name{protest}
\alias{protest}
\title{Have ever joined a protest rally past 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
protest
}
\details{
Question Over the past 5 years have you done any of the following to express your opinion about an issue or your support for a cause?
C. Join a protest rally or march
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2912/vshow}{GSS Data Explorer page for \code{protest}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 84 \tab 1294 \tab 9 \tab 6 \tab 2765 \cr
   Total \tab 1372 \tab 84 \tab 1294 \tab 9 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{boycott}},
\code{\link{conoffcl}},
\code{\link{getmar}},
\code{\link{getmar1}},
\code{\link{havchld}},
\code{\link{havchld1}},
\code{\link{signpet}},
\code{\link{supfam}},
\code{\link{supfam1}}
}
\concept{Adult Transitions}
\keyword{variable}
