% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{protest5}
\alias{protest5}
\title{Allow damage of govt buildings}
\source{
General Social Survey https://gss.norc.org
}
\description{
protest5
}
\details{
Question There are many ways people or organizations can protest against a government action they strongly oppose. Please show which you think should be allowed and which should not be allowed by circling a number after each question.
E. Seriously damaging government buildings.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely allowed
\item \verb{[2]} probably allowed
\item \verb{[3]} probably not allowed
\item \verb{[4]} definitely not allowed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3986/vshow}{GSS Data Explorer page for \code{protest5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 13 \tab 6 \tab 30 \tab 567 \tab 35 \tab 26 \tab 1534 \cr
   1990 \tab 155 \tab 23 \tab 24 \tab 81 \tab 997 \tab 66 \tab 26 \tab 1372 \cr
   Total \tab 1012 \tab 36 \tab 30 \tab 111 \tab 1564 \tab 101 \tab 52 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{obeylaw}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest6}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{ratetone}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
