% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{protest6}
\alias{protest6}
\title{Allow national anti-govt strike}
\source{
General Social Survey https://gss.norc.org
}
\description{
protest6
}
\details{
Question There are many ways people or organizations can protest against a government action they strongly oppose. Please show which you think should be allowed and which should not be allowed by circling a number after each question.
F. Organizing a nationwide strike of all workers against the government.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3987/vshow}{GSS Data Explorer page for \code{protest6}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 58 \tab 66 \tab 119 \tab 368 \tab 41 \tab 25 \tab 1534 \cr
   1990 \tab 155 \tab 104 \tab 182 \tab 224 \tab 589 \tab 91 \tab 27 \tab 1372 \cr
   1996 \tab 1572 \tab 218 \tab 272 \tab 360 \tab 368 \tab 85 \tab 29 \tab 2904 \cr
   2006 \tab 2992 \tab 266 \tab 360 \tab 439 \tab 431 \tab 22 \tab - \tab 4510 \cr
   Total \tab 5576 \tab 646 \tab 880 \tab 1142 \tab 1756 \tab 239 \tab 81 \tab 10320 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{obeylaw}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{ratetone}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
