% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_444.R
\name{protstrs}
\alias{protstrs}
\title{Feel about people protest gov}
\source{
General Social Survey https://gss.norc.org
}
\description{
protstrs
}
\details{
Question What are your personal feelings about...
A. People who organize protests against a government action they strongly oppose.
}
\section{Values}{

\itemize{
\item \code{1} extremely favorable
\item \code{2} favorable
\item \code{3} neither favorable or unfavorable
\item \code{4} unfavorable
\item \code{5} extremely unfavorable
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4085/vshow}{GSS Data Explorer page for \code{protstrs}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{equalize}},
\code{\link{hlthcare}},
\code{\link{racists}},
\code{\link{revoltrs}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
