% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_43.R
\name{prottemp}
\alias{prottemp}
\title{Feeling thermometer:protestants}
\source{
General Social Survey https://gss.norc.org
}
\description{
prottemp
}
\details{
Question I'd like to get your feelings toward groups that are in the news these days. I will use something we call the feeling thermometer, and here is how it works: I'll read the names of a group and I'd like you to rate that group using the feeling thermometer. Ratings between 50 degrees and 100 degrees mean that you feel favorable and warm toward the group. Ratings between 0 degrees and 50 degrees mean that you don't feel favorable toward the group and that you don't care too much for that group. If we come to a group whose name you Don't recognize, you don't need to rate that group. Just tell me and we'll move on to the next one. IF you do recognize the name, but you don't feel particularly warm or cold toward the group, you would rate the group at the 50 degree mark. How would you rate this group using the thermometer?
C. Protestants
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   B/C/- \tab 2004 \cr
   Full \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/375/vshow}{GSS Data Explorer page for \code{prottemp}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 15 \tab 20 \tab 25 \tab 30 \tab 40 \tab 45 \tab 5 \tab 50 \tab 51 \tab 55 \tab 60 \tab 65 \tab 7 \tab 70 \tab 75 \tab 80 \tab 85 \tab 90 \tab 95 \tab don't know \tab no answer \tab 1 \tab 32 \tab 35 \tab 83 \tab 99 \tab 33 \tab 42 \tab 6 \tab 98 \tab 2 \tab 78 \tab Total \cr
   1986 \tab - \tab 10 \tab 2 \tab 228 \tab 8 \tab 3 \tab 1 \tab 15 \tab 45 \tab 4 \tab 1 \tab 232 \tab 10 \tab 4 \tab 223 \tab 4 \tab 1 \tab 268 \tab 19 \tab 43 \tab 249 \tab 22 \tab 5 \tab 54 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1988 \tab - \tab 10 \tab 2 \tab 217 \tab 11 \tab 3 \tab 3 \tab 10 \tab 28 \tab 1 \tab 1 \tab 409 \tab - \tab 1 \tab 135 \tab 2 \tab - \tab 261 \tab 24 \tab 30 \tab 221 \tab 18 \tab 3 \tab 68 \tab 18 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 531 \tab 9 \tab 4 \tab 140 \tab 5 \tab 5 \tab 1 \tab 15 \tab 27 \tab 1 \tab - \tab 216 \tab - \tab 1 \tab 82 \tab 4 \tab - \tab 173 \tab 15 \tab 16 \tab 184 \tab 23 \tab 5 \tab 61 \tab 14 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1537 \cr
   2004 \tab 1939 \tab 3 \tab 2 \tab 116 \tab 3 \tab 2 \tab 5 \tab 10 \tab 14 \tab 2 \tab 5 \tab 321 \tab - \tab 2 \tab 51 \tab 1 \tab 1 \tab 91 \tab 26 \tab 23 \tab 106 \tab 12 \tab 6 \tab 51 \tab 12 \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2812 \cr
   Total \tab 2470 \tab 32 \tab 10 \tab 701 \tab 27 \tab 13 \tab 10 \tab 50 \tab 114 \tab 8 \tab 7 \tab 1178 \tab 10 \tab 8 \tab 491 \tab 11 \tab 2 \tab 793 \tab 84 \tab 112 \tab 760 \tab 75 \tab 19 \tab 234 \tab 63 \tab 5 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7300 \cr
}
}

\seealso{
Other Core: 
\code{\link{contemp}},
\code{\link{libtemp}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}}
}
\concept{Core}
\keyword{variable}
