% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{prouddem}
\alias{prouddem}
\title{How proud the way democracy works}
\source{
General Social Survey https://gss.norc.org
}
\description{
prouddem
}
\details{
Question How proud are you of America in each of the following?
A. The way democracy works
}
\section{Values}{

\itemize{
\item \verb{[1]} very proud
\item \verb{[2]} somewhat proud
\item \verb{[3]} not very proud
\item \verb{[4]} not proud at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4840/vshow}{GSS Data Explorer page for \code{prouddem}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   1996 \tab 1537 \tab 369 \tab 694 \tab 180 \tab 34 \tab 65 \tab 25 \tab - \tab - \tab 2904 \cr
   2004 \tab 1596 \tab 379 \tab 640 \tab 109 \tab 22 \tab 65 \tab 1 \tab - \tab - \tab 2812 \cr
   2014 \tab 1264 \tab 336 \tab 605 \tab 184 \tab 48 \tab 98 \tab 3 \tab - \tab - \tab 2538 \cr
   2024 \tab 1661 \tab 249 \tab 755 \tab 364 \tab 115 \tab 144 \tab 3 \tab - \tab 18 \tab 3309 \cr
   Total \tab 6058 \tab 1333 \tab 2694 \tab 837 \tab 219 \tab 372 \tab 32 \tab 0 \tab 18 \tab 11563 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}

Other United States: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
