% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{proudeco}
\alias{proudeco}
\title{How proud america's economic achievements}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudeco
}
\details{
Question How proud are you of America in each of the following?
C. America''s economic achievements
}
\section{Values}{

\itemize{
\item \code{1} very proud
\item \code{2} somewhat proud
\item \code{3} not very proud
\item \code{4} not proud at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4842/vshow}{GSS Data Explorer page for \code{proudeco}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\keyword{variable}
