% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_307.R
\name{proudemp}
\alias{proudemp}
\title{R proud to work for employer}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudemp
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
O. I am proud to be working for my employer
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2788/vshow}{GSS Data Explorer page for \code{proudemp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 673 \tab 910 \tab 150 \tab 32 \tab 14 \tab 17 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 622 \tab 906 \tab 123 \tab 31 \tab 35 \tab 17 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 427 \tab 614 \tab 84 \tab 25 \tab 10 \tab 27 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 468 \tab 673 \tab 71 \tab 18 \tab 8 \tab 12 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 459 \tab 852 \tab 81 \tab 15 \tab 7 \tab 34 \tab - \tab 2348 \cr
   2022 \tab 1548 \tab 726 \tab 1024 \tab 151 \tab 44 \tab 13 \tab 11 \tab 27 \tab 3544 \cr
   Total \tab 8338 \tab 3375 \tab 4979 \tab 660 \tab 165 \tab 87 \tab 118 \tab 27 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{partteam}},
\code{\link{prodctiv}},
\code{\link{promteok}},
\code{\link{safehlth}},
\code{\link{setthngs}},
\code{\link{toofewwk}},
\code{\link{trdunion}},
\code{\link{wkdecide}},
\code{\link{wksmooth}}
}
\concept{Quality of Working Life}
\keyword{variable}
