% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{proudgrp}
\alias{proudgrp}
\title{How proud its fair and equal treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudgrp
}
\details{
Question How proud are you of America in each of the following?
J. Its fair and equal treatment of all groups in society
}
\section{Values}{

\itemize{
\item \verb{[1]} very proud
\item \verb{[2]} somewhat proud
\item \verb{[3]} not very proud
\item \verb{[4]} not proud at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4849/vshow}{GSS Data Explorer page for \code{proudgrp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 233 \tab 500 \tab 398 \tab 146 \tab 64 \tab 26 \tab 2904 \cr
   2004 \tab 1596 \tab 293 \tab 574 \tab 233 \tab 62 \tab 52 \tab 2 \tab 2812 \cr
   2014 \tab 1264 \tab 216 \tab 488 \tab 347 \tab 128 \tab 92 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 742 \tab 1562 \tab 978 \tab 336 \tab 208 \tab 31 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
