% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{proudhis}
\alias{proudhis}
\title{How proud its history}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudhis
}
\details{
Question How proud are you of America in each of the following?
I. Its history
}
\section{Values}{

\itemize{
\item \code{1} very proud
\item \code{2} somewhat proud
\item \code{3} not very proud
\item \code{4} not proud at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4848/vshow}{GSS Data Explorer page for \code{proudhis}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 638 \tab 491 \tab 108 \tab 51 \tab 52 \tab 27 \tab 2904 \cr
   2004 \tab 1596 \tab 727 \tab 365 \tab 70 \tab 22 \tab 31 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 648 \tab 460 \tab 78 \tab 33 \tab 52 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 2013 \tab 1316 \tab 256 \tab 106 \tab 135 \tab 31 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
