% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{proudsci}
\alias{proudsci}
\title{How proud its scientific and tech achievements}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudsci
}
\details{
Question How proud are you of America in each of the following?
E. Its scientific and technological achievements
}
\section{Values}{

\itemize{
\item \verb{[1]} very proud
\item \verb{[2]} somewhat proud
\item \verb{[3]} not very proud
\item \verb{[4]} not proud at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4844/vshow}{GSS Data Explorer page for \code{proudsci}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 648 \tab 563 \tab 56 \tab 9 \tab 63 \tab 28 \tab 2904 \cr
   2004 \tab 1596 \tab 691 \tab 433 \tab 38 \tab 14 \tab 38 \tab 2 \tab 2812 \cr
   2014 \tab 1264 \tab 552 \tab 565 \tab 53 \tab 20 \tab 81 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 1891 \tab 1561 \tab 147 \tab 43 \tab 182 \tab 33 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\keyword{variable}
