% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{proudsss}
\alias{proudsss}
\title{How proud its social security system}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudsss
}
\details{
Question How proud are you of America in each of the following?
D. Its social security system
}
\section{Values}{

\itemize{
\item \verb{[1]} very proud
\item \verb{[2]} somewhat proud
\item \verb{[3]} not very proud
\item \verb{[4]} not proud at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4843/vshow}{GSS Data Explorer page for \code{proudsss}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 181 \tab 468 \tab 497 \tab 152 \tab 43 \tab 26 \tab 2904 \cr
   2004 \tab 1596 \tab 150 \tab 488 \tab 397 \tab 132 \tab 47 \tab 2 \tab 2812 \cr
   2014 \tab 1264 \tab 161 \tab 459 \tab 390 \tab 174 \tab 86 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 492 \tab 1415 \tab 1284 \tab 458 \tab 176 \tab 32 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}}
}
\concept{ISSP National Identity}
\keyword{variable}
