% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{proz1}
\alias{proz1}
\title{Prozac is harmful to the body}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz1
}
\details{
Question How much do you agree or disagree with the following statements?
A. Prozac is harmful to the body.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2226/vshow}{GSS Data Explorer page for \code{proz1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 59 \tab 240 \tab 371 \tab 211 \tab 28 \tab 279 \tab 9 \tab 2832 \cr
   2006 \tab 3238 \tab 78 \tab 236 \tab 537 \tab 225 \tab 35 \tab 144 \tab 17 \tab 4510 \cr
   Total \tab 4873 \tab 137 \tab 476 \tab 908 \tab 436 \tab 63 \tab 423 \tab 26 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
