% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{proz2}
\alias{proz2}
\title{If no symptoms should discontinue prozac}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz2
}
\details{
Question How much do you agree or disagree with the following statements?
B. If symptoms are no longer present, people should discontinue Prozac.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2227/vshow}{GSS Data Explorer page for \code{proz2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 126 \tab 428 \tab 163 \tab 225 \tab 52 \tab 192 \tab 11 \tab 2832 \cr
   Total \tab 1635 \tab 126 \tab 428 \tab 163 \tab 225 \tab 52 \tab 192 \tab 11 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
