% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{proz3}
\alias{proz3}
\title{Taking prozac interferes with daily activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz3
}
\details{
Question How much do you agree or disagree with the following statements?
C. Taking Prozac interferes with daily activities.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2228/vshow}{GSS Data Explorer page for \code{proz3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 46 \tab 246 \tab 262 \tab 323 \tab 36 \tab 272 \tab 12 \tab 2832 \cr
   2006 \tab 3238 \tab 62 \tab 278 \tab 403 \tab 318 \tab 49 \tab 146 \tab 16 \tab 4510 \cr
   Total \tab 4873 \tab 108 \tab 524 \tab 665 \tab 641 \tab 85 \tab 418 \tab 28 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
