% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{proz5}
\alias{proz5}
\title{Taking prozac makes relations easier}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz5
}
\details{
Question How much do you agree or disagree with the following statements?
E. Taking Prozac makes things easier in relations with family and friends.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2230/vshow}{GSS Data Explorer page for \code{proz5}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 50 \tab 548 \tab 246 \tab 93 \tab 13 \tab 235 \tab 12 \tab 2832 \cr
   2006 \tab 3238 \tab 74 \tab 585 \tab 348 \tab 88 \tab 20 \tab 140 \tab 17 \tab 4510 \cr
   Total \tab 4873 \tab 124 \tab 1133 \tab 594 \tab 181 \tab 33 \tab 375 \tab 29 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
