% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{proz6}
\alias{proz6}
\title{Prozac helps people control their symptoms}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz6
}
\details{
Question How much do you agree or disagree with the following statements?
F. I think Prozac helps people control their symptoms.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2231/vshow}{GSS Data Explorer page for \code{proz6}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 53 \tab 647 \tab 207 \tab 58 \tab 11 \tab 210 \tab 11 \tab 2832 \cr
   2006 \tab 3238 \tab 87 \tab 700 \tab 261 \tab 60 \tab 21 \tab 127 \tab 16 \tab 4510 \cr
   Total \tab 4873 \tab 140 \tab 1347 \tab 468 \tab 118 \tab 32 \tab 337 \tab 27 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
