% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{proz7}
\alias{proz7}
\title{Prozac makes people feel better about themselves}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz7
}
\details{
Question How much do you agree or disagree with the following statements?
G. Taking Prozac makes people feel better about themselves.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2232/vshow}{GSS Data Explorer page for \code{proz7}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 52 \tab 491 \tab 272 \tab 96 \tab 23 \tab 250 \tab 13 \tab 2832 \cr
   2006 \tab 3238 \tab 71 \tab 545 \tab 392 \tab 95 \tab 26 \tab 127 \tab 16 \tab 4510 \cr
   Total \tab 4873 \tab 123 \tab 1036 \tab 664 \tab 191 \tab 49 \tab 377 \tab 29 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
