% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{proz8}
\alias{proz8}
\title{Prozac should be taken to enhance personalities}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz8
}
\details{
Question How much do you agree or disagree with the following statements?
H. Prozac should be taken to enhance the strengths of people''s personalities.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2233/vshow}{GSS Data Explorer page for \code{proz8}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 12 \tab 98 \tab 211 \tab 430 \tab 186 \tab 247 \tab 13 \tab 2832 \cr
   2006 \tab 3238 \tab 15 \tab 110 \tab 288 \tab 473 \tab 234 \tab 136 \tab 16 \tab 4510 \cr
   Total \tab 4873 \tab 27 \tab 208 \tab 499 \tab 903 \tab 420 \tab 383 \tab 29 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
