% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{proz9}
\alias{proz9}
\title{Prozac is for those diagnosed with depression}
\source{
General Social Survey https://gss.norc.org
}
\description{
proz9
}
\details{
Question How much do you agree or disagree with the following statements?
I. Only people who are diagnosed with clinical depression should take Prozac.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2234/vshow}{GSS Data Explorer page for \code{proz9}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 137 \tab 451 \tab 206 \tab 113 \tab 24 \tab 253 \tab 13 \tab 2832 \cr
   2006 \tab 3238 \tab 232 \tab 494 \tab 269 \tab 90 \tab 25 \tab 146 \tab 16 \tab 4510 \cr
   Total \tab 4873 \tab 369 \tab 945 \tab 475 \tab 203 \tab 49 \tab 399 \tab 29 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
