% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{prozfor1}
\alias{prozfor1}
\title{What can you tell me about prozac is used for}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozfor1
}
\details{
Question What can you tell me about what Prozac is used for?
A. First mention
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2223/vshow}{GSS Data Explorer page for \code{prozfor1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 3 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab 10 \tab 27 \tab 28 \tab 29 \tab 30 \tab 32 \tab 34 \tab 95 \tab 97 \tab Total \cr
   1998 \tab 1635 \tab 136 \tab 23 \tab 3 \tab 8 \tab 1 \tab 16 \tab 7 \tab 11 \tab 2 \tab 4 \tab 539 \tab 24 \tab 1 \tab 5 \tab 5 \tab 4 \tab 13 \tab 1 \tab 34 \tab 57 \tab 18 \tab 13 \tab 16 \tab 60 \tab 4 \tab 14 \tab 144 \tab 34 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2006 \tab 3252 \tab 711 \tab 9 \tab 26 \tab 23 \tab 25 \tab 1 \tab 8 \tab - \tab 51 \tab - \tab - \tab 9 \tab 11 \tab - \tab - \tab 9 \tab 5 \tab 18 \tab 1 \tab - \tab - \tab - \tab 31 \tab - \tab - \tab 1 \tab 1 \tab 10 \tab 90 \tab 12 \tab 1 \tab 2 \tab 8 \tab 1 \tab 6 \tab 91 \tab 97 \tab 4510 \cr
   Total \tab 4887 \tab 847 \tab 32 \tab 29 \tab 31 \tab 26 \tab 17 \tab 15 \tab 11 \tab 53 \tab 4 \tab 539 \tab 33 \tab 12 \tab 5 \tab 5 \tab 13 \tab 18 \tab 19 \tab 35 \tab 57 \tab 18 \tab 13 \tab 47 \tab 60 \tab 4 \tab 15 \tab 145 \tab 44 \tab 90 \tab 12 \tab 1 \tab 2 \tab 8 \tab 1 \tab 6 \tab 91 \tab 97 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
