% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{prozfor3}
\alias{prozfor3}
\title{What can you tell me about prozac is used for}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozfor3
}
\details{
Question What can you tell me about what Prozac is used for?
C. Third Mention
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2225/vshow}{GSS Data Explorer page for \code{prozfor3}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 19 \tab 2 \tab 20 \tab 21 \tab 24 \tab 25 \tab 3 \tab 4 \tab 5 \tab 50 \tab 6 \tab 8 \tab no answer \tab 10 \tab 18 \tab 26 \tab 30 \tab 31 \tab 32 \tab 34 \tab don't know \tab Total \cr
   1998 \tab 2757 \tab 9 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 8 \tab 1 \tab 1 \tab 7 \tab 6 \tab 3 \tab 6 \tab 3 \tab 5 \tab 3 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2006 \tab 3252 \tab - \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 5 \tab 2 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 1216 \tab 4510 \cr
   Total \tab 6009 \tab 9 \tab 4 \tab 4 \tab 4 \tab 5 \tab 1 \tab 7 \tab 5 \tab 8 \tab 5 \tab 3 \tab 10 \tab 6 \tab 3 \tab 6 \tab 3 \tab 5 \tab 7 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 1216 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
