% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{prozkid1}
\alias{prozkid1}
\title{Give kid prozac if is hostile}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozkid1
}
\details{
Question How likely would you be to give doctor prescribed Prozac to your child or a child you were responsible for in the following situations (A child would be considered anyone between the age of 8 and 15).
A. ...because s/he is hostile, often loses his/her temper, often argues with adults, actively defies authority and seems spiteful or vindictive.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2239/vshow}{GSS Data Explorer page for \code{prozkid1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 67 \tab 160 \tab 132 \tab 173 \tab 525 \tab 127 \tab 13 \tab 2832 \cr
   Total \tab 1635 \tab 67 \tab 160 \tab 132 \tab 173 \tab 525 \tab 127 \tab 13 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
