% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{prozkid3}
\alias{prozkid3}
\title{Give kid prozac if is talking about suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozkid3
}
\details{
Question How likely would you be to give doctor prescribed Prozac to your child or a child you were responsible for in the following situations (A child would be considered anyone between the age of 8 and 15).
B. ...because s/he is talking about killing him or herself.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2241/vshow}{GSS Data Explorer page for \code{prozkid3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 237 \tab 243 \tab 180 \tab 78 \tab 310 \tab 136 \tab 13 \tab 2832 \cr
   Total \tab 1635 \tab 237 \tab 243 \tab 180 \tab 78 \tab 310 \tab 136 \tab 13 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
