% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{prozslf1}
\alias{prozslf1}
\title{Take prozac if trouble in personal life}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozslf1
}
\details{
Question How likely would you be to take doctor prescribed Prozac in the following situations?
A. ...because you were having trouble in your personal life.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} mixed
\item \code{4} somewht unlikely
\item \code{5} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2235/vshow}{GSS Data Explorer page for \code{prozslf1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 59 \tab 117 \tab 95 \tab 191 \tab 634 \tab 91 \tab 10 \tab 2832 \cr
   2006 \tab 3238 \tab 67 \tab 141 \tab 196 \tab 234 \tab 549 \tab 68 \tab 17 \tab 4510 \cr
   Total \tab 4873 \tab 126 \tab 258 \tab 291 \tab 425 \tab 1183 \tab 159 \tab 27 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
