% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{prozslf2}
\alias{prozslf2}
\title{Take prozac if cant cope with stresses}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozslf2
}
\details{
Question How likely would you be to take doctor prescribed Prozac in the following situations?
B. ...because you didn''t know how to cope anymore with the stresses of life.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2236/vshow}{GSS Data Explorer page for \code{prozslf2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 104 \tab 210 \tab 136 \tab 191 \tab 444 \tab 100 \tab 12 \tab 2832 \cr
   2006 \tab 3238 \tab 124 \tab 283 \tab 201 \tab 190 \tab 389 \tab 68 \tab 17 \tab 4510 \cr
   Total \tab 4873 \tab 228 \tab 493 \tab 337 \tab 381 \tab 833 \tab 168 \tab 29 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf3}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
