% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{prozslf3}
\alias{prozslf3}
\title{Take prozac if feel depressed sleepless}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozslf3
}
\details{
Question How likely would you be to take doctor prescribed Prozac in the following situations?
C. ...because you were feeling depressed, tired, were having trouble sleeping and concentrating, and felt worthless.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2237/vshow}{GSS Data Explorer page for \code{prozslf3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 127 \tab 234 \tab 126 \tab 176 \tab 428 \tab 95 \tab 11 \tab 2832 \cr
   2006 \tab 3238 \tab 159 \tab 300 \tab 182 \tab 162 \tab 387 \tab 65 \tab 17 \tab 4510 \cr
   Total \tab 4873 \tab 286 \tab 534 \tab 308 \tab 338 \tab 815 \tab 160 \tab 28 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
