% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{prozslf4}
\alias{prozslf4}
\title{Take prozac if intense fear losing control}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozslf4
}
\details{
Question How likely would you be to take doctor prescribed Prozac in the following situations?
D. ...for no apparent reason, you were having periods of intense fear in which you were trembling, sweating, feeling dizzy, and feared losing control or going crazy.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2238/vshow}{GSS Data Explorer page for \code{prozslf4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 186 \tab 292 \tab 152 \tab 126 \tab 319 \tab 109 \tab 13 \tab 2832 \cr
   2006 \tab 3238 \tab 244 \tab 312 \tab 166 \tab 144 \tab 322 \tab 66 \tab 18 \tab 4510 \cr
   Total \tab 4873 \tab 430 \tab 604 \tab 318 \tab 270 \tab 641 \tab 175 \tab 31 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}}
}
\concept{Mental Health}
\keyword{variable}
