% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{prvdschl}
\alias{prvdschl}
\title{Who provides school educ}
\source{
General Social Survey https://gss.norc.org
}
\description{
prvdschl
}
\details{
Question (HAND CARD C3) 
(People have different opinions on who should provide services in America. Who do you think should primarily provide...)
School education for children?  (Should it be?)
}
\section{Values}{

\itemize{
\item \verb{[1]} government
\item \verb{[2]} private companies/for-profit organizations
\item \verb{[3]} non-profit organizations/charities/cooperatives
\item \verb{[4]} religious organizations
\item \verb{[5]} family, relatives or friends
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6307/vshow}{GSS Data Explorer page for \code{prvdschl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab government \tab private companies/for-profit organizations \tab non-profit organizations/charities/cooperatives \tab religious organizations \tab family, relatives or friends \tab can't choose \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 1071 \tab 76 \tab 63 \tab 8 \tab 131 \tab 25 \tab 16 \tab 2867 \cr
   Total \tab 1477 \tab 1071 \tab 76 \tab 63 \tab 8 \tab 131 \tab 25 \tab 16 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{polgbeco}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
