% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{pryntfam}
\alias{pryntfam}
\title{Med just prevent family from working out problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
pryntfam
}
\details{
Question Now, I'd like to ask you for your opinions on several issues. For each question, please tell me if you strongly agree, agree somewhat, disagree somewhat or strongly disagree with the statement.
G. Medications for behavior problems just prevent families from working out problems themselves.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2967/vshow}{GSS Data Explorer page for \code{pryntfam}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree somewhat \tab disagree somewhat \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 251 \tab 475 \tab 430 \tab 161 \tab 63 \tab 13 \tab 2765 \cr
   Total \tab 1372 \tab 251 \tab 475 \tab 430 \tab 161 \tab 63 \tab 13 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{mntlrel}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
