% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_243.R
\name{psycfdoc}
\alias{psycfdoc}
\title{Would trust family physician prescription}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycfdoc
}
\details{
Question If you were to seek services for a problem with your emotions, nerves, or your mental health, how much would you agree or disagree with the following options?
B. I would trust my family physician to prescribe psychiatric medication.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2183/vshow}{GSS Data Explorer page for \code{psycfdoc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 106 \tab 381 \tab 221 \tab 416 \tab 189 \tab 67 \tab 7 \tab 2832 \cr
   Total \tab 1445 \tab 106 \tab 381 \tab 221 \tab 416 \tab 189 \tab 67 \tab 7 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{mntlcare}},
\code{\link{psycdrct}},
\code{\link{psyckid3}},
\code{\link{psycpsyc}},
\code{\link{swithlth}},
\code{\link{usedmntl}}
}
\concept{Mental Health}
\keyword{variable}
