% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_242.R
\name{psyckid2}
\alias{psyckid2}
\title{Will give kid med because school trouble}
\source{
General Social Survey https://gss.norc.org
}
\description{
psyckid2
}
\details{
Question How likely would you be to give doctor-prescribed psychiatric medication to your child or a child you were responsible for in the following situations (A child would be considered anyone between the ages of 8 and 15).
B...because s/he is not paying attention at school, does not follow through with school work and chores, has difficulty organizing activities, is easily distracted, talks excessively, and seems to run around or fidget constantly.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2180/vshow}{GSS Data Explorer page for \code{psyckid2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 126 \tab 250 \tab 222 \tab 259 \tab 444 \tab 82 \tab 4 \tab 2832 \cr
   Total \tab 1445 \tab 126 \tab 250 \tab 222 \tab 259 \tab 444 \tab 82 \tab 4 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}}
}
\concept{Mental Health}
\keyword{variable}
