% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_243.R
\name{psyckid3}
\alias{psyckid3}
\title{Will give kid med because kid talked suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
psyckid3
}
\details{
Question How likely would you be to give doctor-prescribed psychiatric medication to your child or a child you were responsible for in the following situations (A child would be considered anyone between the ages of 8 and 15).
C....because s/he was talking about killing him or herself.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} mixed
\item \code{4} somewht unlikely
\item \code{5} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2181/vshow}{GSS Data Explorer page for \code{psyckid3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 454 \tab 280 \tab 225 \tab 99 \tab 243 \tab 81 \tab 5 \tab 2832 \cr
   Total \tab 1445 \tab 454 \tab 280 \tab 225 \tab 99 \tab 243 \tab 81 \tab 5 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{mntlcare}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psycpsyc}},
\code{\link{swithlth}},
\code{\link{usedmntl}}
}
\concept{Mental Health}
\keyword{variable}
