% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{psycmed1}
\alias{psycmed1}
\title{Psychiatric medicine is harmful to the body}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycmed1
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors to help people who are having problems with their emotions, nerves or their mental health:
A. Psychiatric medicine is harmful to the body.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2168/vshow}{GSS Data Explorer page for \code{psycmed1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 72 \tab 253 \tab 329 \tab 431 \tab 127 \tab 174 \tab 1 \tab 2832 \cr
   2006 \tab 3073 \tab 94 \tab 297 \tab 454 \tab 380 \tab 111 \tab 86 \tab 15 \tab 4510 \cr
   Total \tab 4518 \tab 166 \tab 550 \tab 783 \tab 811 \tab 238 \tab 260 \tab 16 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psycmed2}},
\code{\link{psycmed3}},
\code{\link{seenmntl}}
}
\concept{Mental Health}
\keyword{variable}
