% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_242.R
\name{psycmed7}
\alias{psycmed7}
\title{Helps people feel better about themselves}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycmed7
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors to help people who are having problems with their emotions, nerves or their mental health:
G. Taking medication helps people feel better about themselves.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2174/vshow}{GSS Data Explorer page for \code{psycmed7}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 111 \tab 650 \tab 286 \tab 177 \tab 27 \tab 133 \tab 3 \tab 2832 \cr
   2006 \tab 3073 \tab 147 \tab 766 \tab 279 \tab 132 \tab 24 \tab 75 \tab 14 \tab 4510 \cr
   Total \tab 4518 \tab 258 \tab 1416 \tab 565 \tab 309 \tab 51 \tab 208 \tab 17 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}}
}
\concept{Mental Health}
\keyword{variable}
