% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_243.R
\name{psycpsyc}
\alias{psycpsyc}
\title{Only trust psychiatrist prescription}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycpsyc
}
\details{
Question If you were to seek services for a problem with your emotions, nerves, or your mental health, how much would you agree or disagree with the following options?
C. I would only trust a psychiatrist to prescribe psychiatric medication.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2184/vshow}{GSS Data Explorer page for \code{psycpsyc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 199 \tab 494 \tab 241 \tab 305 \tab 69 \tab 72 \tab 7 \tab 2832 \cr
   Total \tab 1445 \tab 199 \tab 494 \tab 241 \tab 305 \tab 69 \tab 72 \tab 7 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{mntlcare}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid3}},
\code{\link{swithlth}},
\code{\link{usedmntl}}
}
\concept{Mental Health}
\keyword{variable}
