% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{psysamephys}
\alias{psysamephys}
\title{Management supports psychological health}
\source{
General Social Survey https://gss.norc.org
}
\description{
psysamephys
}
\details{
Question (To what extent to do you agree or disagree with the following statements?) Senior management considers psychological health to be as important as productivity.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly disagree
\item \verb{[2]} disagree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} agree
\item \verb{[5]} strongly agree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7857/vshow}{GSS Data Explorer page for \code{psysamephys}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly disagree \tab disagree \tab neither agree nor disagree \tab agree \tab strongly agree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1549 \tab 245 \tab 262 \tab 510 \tab 684 \tab 218 \tab 28 \tab 26 \tab 22 \tab 3544 \cr
   Total \tab 1549 \tab 245 \tab 262 \tab 510 \tab 684 \tab 218 \tab 28 \tab 26 \tab 22 \tab 3544 \cr
}
}

\keyword{variable}
