% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{ptnrsxnow1}
\alias{ptnrsxnow1}
\title{Recent partner's sex now (3 category)}
\source{
General Social Survey https://gss.norc.org
}
\description{
ptnrsxnow1
}
\details{
Question To the best of your knowledge, does your most recent sexual partner describe themselves as male, female, or (REDACTED)?
Question edits are in compliance with federal administration directives released in January 2025.
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[3]} redacted/none of these
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8169/vshow}{GSS Data Explorer page for \code{ptnrsxnow1}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab male \tab female \tab redacted/none of these \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1062 \tab 836 \tab 22 \tab 59 \tab 1273 \tab 8 \tab 49 \tab 3309 \cr
   Total \tab 0 \tab 1062 \tab 836 \tab 22 \tab 59 \tab 1273 \tab 8 \tab 49 \tab 3309 \cr
}
}

\concept{High Risk Behaviors}
\concept{Sex}
\keyword{variable}
