% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{pubagncy}
\alias{pubagncy}
\title{Registered at a public employment agency}
\source{
General Social Survey https://gss.norc.org
}
\description{
pubagncy
}
\details{
Question Thinking about the last 12 months, have you done any of the following in order to find a job?
A. Registered at a public employment agency?
}
\section{Values}{

\itemize{
\item \verb{[1]} no
\item \verb{[2]} yes, once or twice
\item \verb{[3]} yes, more than twice
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4599/vshow}{GSS Data Explorer page for \code{pubagncy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab yes, once or twice \tab yes, more than twice \tab no answer \tab Total \cr
   1998 \tab 2460 \tab 274 \tab 24 \tab 7 \tab 67 \tab 2832 \cr
   2006 \tab 4008 \tab 448 \tab 41 \tab 9 \tab 4 \tab 4510 \cr
   2016 \tab 2533 \tab 273 \tab 28 \tab 10 \tab 23 \tab 2867 \cr
   Total \tab 9001 \tab 995 \tab 93 \tab 26 \tab 94 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
