% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{pubsch}
\alias{pubsch}
\title{All children attend pub elem & hi school}
\source{
General Social Survey https://gss.norc.org
}
\description{
pubsch
}
\details{
Question Do/did all of your children attend public school for all of their elementary and high school education?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1147/vshow}{GSS Data Explorer page for \code{pubsch}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1998 \tab 1940 \tab 694 \tab 161 \tab 37 \tab - \tab 2832 \cr
   2000 \tab 1093 \tab 1445 \tab 258 \tab 19 \tab 2 \tab 2817 \cr
   Total \tab 3033 \tab 2139 \tab 419 \tab 56 \tab 2 \tab 5649 \cr
}
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{givearts}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{kid5up}},
\code{\link{numcong}},
\code{\link{relhrs1}},
\code{\link{relhrs2}}
}
\concept{Religion}
\keyword{variable}
