% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{punarms}
\alias{punarms}
\title{Punishment stealing and selling arms}
\source{
General Social Survey https://gss.norc.org
}
\description{
punarms
}
\details{
Question I'm going to ask you about several offenses. For each case, indicate what you would consider to be the appropriate punishment when the offender is a military service member or government employee.
A. Stealing and selling weapons, ammunition, and explosives from a military depot.
}
\section{Values}{

\itemize{
\item \verb{[1]} reprimand
\item \verb{[2]} dismissal from the military/fired by the government
\item \verb{[3]} one year in prison
\item \verb{[4]} 5-9 years in prison
\item \verb{[5]} 10-20 years in prison
\item \verb{[6]} life in prison
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5232/vshow}{GSS Data Explorer page for \code{punarms}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab reprimand \tab dismissal from the military/fired by the government \tab one year in prison \tab 5-9 years in prison \tab 10-20 years in prison \tab life in prison \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 19 \tab 271 \tab 100 \tab 297 \tab 412 \tab 246 \tab 56 \tab 18 \tab 2817 \cr
   Total \tab 1398 \tab 19 \tab 271 \tab 100 \tab 297 \tab 412 \tab 246 \tab 56 \tab 18 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}},
\code{\link{taketrck}}
}
\concept{National Security}
\keyword{variable}
