% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{punsin}
\alias{punsin}
\title{R agrees that sinners must be punished}
\source{
General Social Survey https://gss.norc.org
}
\description{
punsin
}
\details{
Question Please consider the following statements and tell me whether you agree strongly, agree somewhat, disagree somewhat or disagree strongly with each statement.
A. Those who violate God''s rules must be punished.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} disagree somewhat
\item \verb{[4]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1081/vshow}{GSS Data Explorer page for \code{punsin}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 324 \tab 551 \tab 362 \tab 147 \tab 87 \tab 10 \tab 1481 \cr
   2006 \tab 1518 \tab 433 \tab 824 \tab 860 \tab 680 \tab 167 \tab 28 \tab 4510 \cr
   2008 \tab - \tab 299 \tab 559 \tab 566 \tab 457 \tab 105 \tab 37 \tab 2023 \cr
   2010 \tab - \tab 296 \tab 515 \tab 590 \tab 521 \tab 84 \tab 38 \tab 2044 \cr
   Total \tab 1518 \tab 1352 \tab 2449 \tab 2378 \tab 1805 \tab 443 \tab 113 \tab 10058 \cr
}
}

\seealso{
Other Religion: 
\code{\link{blkwhite}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{gochurch}},
\code{\link{permoral}},
\code{\link{readword}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
