% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_142.R
\name{quality}
\alias{quality}
\title{How easy your supervisor know how good you are}
\source{
General Social Survey https://gss.norc.org
}
\description{
quality
}
\details{
Question How easy or hard would you say it is for you supervisor to evaluate the quality of the work that is done by a person in a job like yours?  Would you say very easy, somewhat easy, somewhat hard, or very hard?
}
\section{Values}{

\itemize{
\item \verb{[1]} very easy
\item \verb{[2]} somewhat easy
\item \verb{[3]} somewhat hard
\item \verb{[4]} very hard
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1337/vshow}{GSS Data Explorer page for \code{quality}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very easy \tab somewhat easy \tab somewhat hard \tab very hard \tab don't know \tab no answer \tab Total \cr
   1991 \tab 776 \tab 383 \tab 201 \tab 102 \tab 34 \tab 3 \tab 18 \tab 1517 \cr
   Total \tab 776 \tab 383 \tab 201 \tab 102 \tab 34 \tab 3 \tab 18 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{quantity}},
\code{\link{totmoney}},
\code{\link{wrkwell}},
\code{\link{youmoney}}
}
\concept{Work Organizations}
\keyword{variable}
