% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_142.R
\name{quantity}
\alias{quantity}
\title{How easy your supervisor know how much you do}
\source{
General Social Survey https://gss.norc.org
}
\description{
quantity
}
\details{
Question How easy or hard would you say it is for your supervisor to know how much work a person in a job like yours does in a week?  Is it very easy, somewhat easy, somewhat hard or very hard?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} somewhat easy
\item \code{3} somewhat hard
\item \code{4} very hard
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1336/vshow}{GSS Data Explorer page for \code{quantity}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very easy \tab somewhat easy \tab somewhat hard \tab very hard \tab don't know \tab no answer \tab Total \cr
   1991 \tab 795 \tab 362 \tab 179 \tab 102 \tab 60 \tab 4 \tab 15 \tab 1517 \cr
   Total \tab 795 \tab 362 \tab 179 \tab 102 \tab 60 \tab 4 \tab 15 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{quality}},
\code{\link{totmoney}},
\code{\link{wrkwell}},
\code{\link{youmoney}}
}
\concept{Work Organizations}
\keyword{variable}
