% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racavoid}
\alias{racavoid}
\title{Avoid driving thru black neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
racavoid
}
\details{
Question If you were driving through the neighborhoods in a city, would you go out of your way of your way to avoid going through a (Negro/Black/African-American) section?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/725/vshow}{GSS Data Explorer page for \code{racavoid}}.
}

\seealso{
Other Core: 
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racquit}},
\code{\link{racsups}},
\code{\link{racteach}}

Other Race: 
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racquit}},
\code{\link{racsups}},
\code{\link{racteach}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
