% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{raccare}
\alias{raccare}
\title{R's concern about racial issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
raccare
}
\details{
Question How concerned are you personally about race relations?  Are you very concerned, somewhat concerned, not very concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \code{1} very concerned
\item \code{2} somewhat concerned
\item \code{3} not very concerned
\item \code{4} not concerned at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/418/vshow}{GSS Data Explorer page for \code{raccare}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very concerned \tab somewhat concerned \tab not very concerned \tab not concerned at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab 804 \tab 153 \tab 388 \tab 173 \tab 71 \tab 6 \tab 4 \tab 1599 \cr
   Total \tab 804 \tab 153 \tab 388 \tab 173 \tab 71 \tab 6 \tab 4 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{blksimp}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}
}
\concept{Core}
\keyword{variable}
