% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racchng}
\alias{racchng}
\title{Try to integrate social club}
\source{
General Social Survey https://gss.norc.org
}
\description{
racchng
}
\details{
Question If you and your friends belonged to a social club that would not let (Whites/Negroes/Blacks) join, would you try to change the rules so that (Whites/(Negroes/Blacks/African-Americans) could join?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} wdnt belong
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/726/vshow}{GSS Data Explorer page for \code{racchng}}.
}

\seealso{
Other Core: 
\code{\link{racavoid}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racquit}},
\code{\link{racsups}},
\code{\link{racteach}}
}
\concept{Core}
\keyword{variable}
