% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{racdif1x}
\alias{racdif1x}
\title{Differences due to discrimination (longer intro)}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif1x
}
\details{
Question
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7963/vshow}{GSS Data Explorer page for \code{racdif1x}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 672 \tab 484 \tab 42 \tab 2339 \tab 5 \tab 2 \tab 3544 \cr
   2024 \tab - \tab 564 \tab 505 \tab 29 \tab 2200 \tab 1 \tab 10 \tab 3309 \cr
   Total \tab 0 \tab 1236 \tab 989 \tab 71 \tab 4539 \tab 6 \tab 12 \tab 6853 \cr
}
}

\keyword{variable}
