% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_664.R
\name{racdif3x}
\alias{racdif3x}
\title{Differences due to lack of education}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif3x
}
\details{
Question
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7965/vshow}{GSS Data Explorer page for \code{racdif3x}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab skipped on web \tab no answer \tab Total \cr
   2022 \tab - \tab 652 \tab 518 \tab 31 \tab 2339 \tab 4 \tab - \tab 3544 \cr
   2024 \tab - \tab 554 \tab 518 \tab 25 \tab 2200 \tab 8 \tab 4 \tab 3309 \cr
   Total \tab 0 \tab 1206 \tab 1036 \tab 56 \tab 4539 \tab 12 \tab 4 \tab 6853 \cr
}
}

\keyword{variable}
