% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{racdif4}
\alias{racdif4}
\title{Differences due to lack of will}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif4
}
\details{
Question On the average (Negroes/Blacks/African-Americans) have worse jobs, income, and housing than white people.Do you think these differences are . . .
D. Because most (Negroes/Blacks/African-Americans) just don't have the motivation or will power to pull themselves up out of poverty?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1988, 1989, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C \tab 1990 \cr
   Full \tab 1977, 1985, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/731/vshow}{GSS Data Explorer page for \code{racdif4}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1977 \tab 176 \tab 836 \tab 434 \tab 77 \tab 7 \tab - \tab 1530 \cr
   1985 \tab - \tab 846 \tab 609 \tab 76 \tab 3 \tab - \tab 1534 \cr
   1986 \tab - \tab 857 \tab 547 \tab 60 \tab 6 \tab - \tab 1470 \cr
   1988 \tab 493 \tab 539 \tab 393 \tab 49 \tab 7 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 565 \tab 378 \tab 55 \tab 8 \tab - \tab 1537 \cr
   1990 \tab - \tab 788 \tab 493 \tab 83 \tab 8 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 562 \tab 399 \tab 59 \tab 4 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 527 \tab 476 \tab 72 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 964 \tab 882 \tab 115 \tab 16 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 926 \tab 902 \tab 120 \tab 12 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 790 \tab 915 \tab 157 \tab 9 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 845 \tab 884 \tab 130 \tab 18 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 435 \tab 425 \tab 45 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 426 \tab 431 \tab 42 \tab 7 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 937 \tab 955 \tab 92 \tab 8 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 651 \tab 615 \tab 58 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 652 \tab 705 \tab 61 \tab 12 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 619 \tab 619 \tab 54 \tab 10 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 730 \tab 890 \tab 45 \tab 10 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 786 \tab 1035 \tab 62 \tab 5 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 566 \tab 927 \tab 55 \tab 11 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 706 \tab 1960 \tab 14 \tab 1 \tab 36 \tab 4032 \cr
   2022 \tab 2373 \tab 348 \tab 769 \tab 43 \tab 5 \tab 6 \tab 3544 \cr
   Total \tab 20659 \tab 15901 \tab 16643 \tab 1624 \tab 180 \tab 42 \tab 55049 \cr
}
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{unemp5}}
}
\concept{Core}
\keyword{variable}
