% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_664.R
\name{racdif4x}
\alias{racdif4x}
\title{Differences due to lack of will}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif4x
}
\details{
Question
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7966/vshow}{GSS Data Explorer page for \code{racdif4x}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab skipped on web \tab no answer \tab Total \cr
   2022 \tab - \tab 342 \tab 822 \tab 37 \tab 2339 \tab 4 \tab - \tab 3544 \cr
   2024 \tab - \tab 322 \tab 738 \tab 37 \tab 2200 \tab 10 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 664 \tab 1560 \tab 74 \tab 4539 \tab 14 \tab 2 \tab 6853 \cr
}
}

\keyword{variable}
