% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{racdif5}
\alias{racdif5}
\title{Differences due mainly to unbringing}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif5
}
\details{
Question On the average Black people have worse jobs, income, and housing than white people.Do you think these differences are . . .
E. Mainly due to upbringing?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8202/vshow}{GSS Data Explorer page for \code{racdif5}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 631 \tab 383 \tab 43 \tab 2235 \tab 3 \tab 14 \tab 3309 \cr
   Total \tab 0 \tab 631 \tab 383 \tab 43 \tab 2235 \tab 3 \tab 14 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conlabory}},
\code{\link{marsamey}},
\code{\link{norelgsp16}},
\code{\link{pawrkgrw}}
}
\concept{Core}
\keyword{variable}
