% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{racdif5x}
\alias{racdif5x}
\title{Differences due mainly to upbringing (longer intro)}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif5x
}
\details{
Question On the average Black people have worse jobs, income, and housing than white people.Do you think these differences are . . .
E. Mainly due to upbringing?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8203/vshow}{GSS Data Explorer page for \code{racdif5x}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 680 \tab 377 \tab 39 \tab 2200 \tab 5 \tab 8 \tab 3309 \cr
   Total \tab 0 \tab 680 \tab 377 \tab 39 \tab 2200 \tab 5 \tab 8 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conorelgsp}},
\code{\link{family16sex}},
\code{\link{femsee1}},
\code{\link{mascsee1}},
\code{\link{sexnow2}}
}
\concept{Core}
\keyword{variable}
