% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{racdin}
\alias{racdin}
\title{Object to opp. race home for dinner}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdin
}
\details{
Question How strongly would you object if a member of your family wanted to bring a (Negro/Black) friend home to dinner?
Would you object strongly, mildly, or not at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly object
\item \verb{[2]} mildly object
\item \verb{[3]} not object
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/384/vshow}{GSS Data Explorer page for \code{racdin}}.

Counts by year:\tabular{llllllll}{
   year \tab strongly object \tab mildly object \tab not object \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 173 \tab 206 \tab 939 \tab 30 \tab 261 \tab 4 \tab 1613 \cr
   1973 \tab 201 \tab 196 \tab 895 \tab 21 \tab 183 \tab 8 \tab 1504 \cr
   1974 \tab 141 \tab 204 \tab 943 \tab 22 \tab 173 \tab 1 \tab 1484 \cr
   1976 \tab 173 \tab 196 \tab 963 \tab 33 \tab 129 \tab 5 \tab 1499 \cr
   1977 \tab 154 \tab 220 \tab 959 \tab 16 \tab 176 \tab 5 \tab 1530 \cr
   1980 \tab 149 \tab 190 \tab 1105 \tab 21 \tab - \tab 3 \tab 1468 \cr
   1982 \tab 139 \tab 186 \tab 1508 \tab 21 \tab - \tab 6 \tab 1860 \cr
   1984 \tab 93 \tab 160 \tab 1197 \tab 18 \tab - \tab 5 \tab 1473 \cr
   1985 \tab 73 \tab 79 \tab 591 \tab 3 \tab 783 \tab 5 \tab 1534 \cr
   Total \tab 1296 \tab 1637 \tab 9100 \tab 185 \tab 1705 \tab 42 \tab 13965 \cr
}
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}},
\code{\link{relgrade}}
}
\concept{Core}
\keyword{variable}
