% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{race}
\alias{race}
\title{Race of respondent}
\source{
General Social Survey https://gss.norc.org
}
\description{
race
}
\details{
Question CODE WITHOUT ASKING ONLY IF THERE IS NO DOUBT IN YOUR MIND.
What race do you consider yourself? RECORD VERBATIM AND CODE.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black
\item \code{3} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/82/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllll}{
   year \tab black \tab other \tab white \tab iap \tab Total \cr
   1972 \tab 261 \tab 4 \tab 1348 \tab - \tab 1613 \cr
   1973 \tab 183 \tab 13 \tab 1308 \tab - \tab 1504 \cr
   1974 \tab 173 \tab 7 \tab 1304 \tab - \tab 1484 \cr
   1975 \tab 163 \tab 4 \tab 1323 \tab - \tab 1490 \cr
   1976 \tab 129 \tab 9 \tab 1361 \tab - \tab 1499 \cr
   1977 \tab 176 \tab 15 \tab 1339 \tab - \tab 1530 \cr
   1978 \tab 158 \tab 16 \tab 1358 \tab - \tab 1532 \cr
   1980 \tab 140 \tab 10 \tab 1318 \tab - \tab 1468 \cr
   1982 \tab 510 \tab 27 \tab 1323 \tab - \tab 1860 \cr
   1983 \tab 165 \tab 18 \tab 1416 \tab - \tab 1599 \cr
   1984 \tab 170 \tab 52 \tab 1251 \tab - \tab 1473 \cr
   1985 \tab 152 \tab 44 \tab 1338 \tab - \tab 1534 \cr
   1986 \tab 184 \tab 37 \tab 1249 \tab - \tab 1470 \cr
   1987 \tab 544 \tab 53 \tab 1222 \tab - \tab 1819 \cr
   1988 \tab 186 \tab 61 \tab 1234 \tab - \tab 1481 \cr
   1989 \tab 157 \tab 61 \tab 1319 \tab - \tab 1537 \cr
   1990 \tab 159 \tab 63 \tab 1150 \tab - \tab 1372 \cr
   1991 \tab 204 \tab 49 \tab 1264 \tab - \tab 1517 \cr
   1993 \tab 179 \tab 80 \tab 1347 \tab - \tab 1606 \cr
   1994 \tab 388 \tab 121 \tab 2483 \tab - \tab 2992 \cr
   1996 \tab 402 \tab 153 \tab 2349 \tab - \tab 2904 \cr
   1998 \tab 400 \tab 191 \tab 2241 \tab - \tab 2832 \cr
   2000 \tab 429 \tab 175 \tab 2213 \tab - \tab 2817 \cr
   2002 \tab 410 \tab 167 \tab 2188 \tab - \tab 2765 \cr
   2004 \tab 377 \tab 201 \tab 2234 \tab - \tab 2812 \cr
   2006 \tab 634 \tab 592 \tab 3284 \tab - \tab 4510 \cr
   2008 \tab 281 \tab 183 \tab 1559 \tab - \tab 2023 \cr
   2010 \tab 311 \tab 183 \tab 1550 \tab - \tab 2044 \cr
   2012 \tab 301 \tab 196 \tab 1477 \tab - \tab 1974 \cr
   2014 \tab 386 \tab 262 \tab 1890 \tab - \tab 2538 \cr
   2016 \tab 490 \tab 277 \tab 2100 \tab - \tab 2867 \cr
   2018 \tab 385 \tab 270 \tab 1693 \tab - \tab 2348 \cr
   2021 \tab 463 \tab 405 \tab 3110 \tab 54 \tab 4032 \cr
   2022 \tab 565 \tab 412 \tab 2514 \tab 53 \tab 3544 \cr
   2024 \tab 572 \tab 392 \tab 2280 \tab 65 \tab 3309 \cr
   Total \tab 10787 \tab 4803 \tab 59937 \tab 172 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
