% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_96.R
\name{race4}
\alias{race4}
\title{Race of person number 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
race4
}
\details{
Question Is (NAME) Asian, Black, Hispanic, White or something else?  ASK FOR EACH NAME.
4. Race of Fourth Person
}
\section{Values}{

\itemize{
\item \code{1} asian
\item \code{2} black
\item \code{3} hispanic
\item \code{4} white
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/873/vshow}{GSS Data Explorer page for \code{race4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab asian \tab black \tab hispanic \tab white \tab other \tab no answer \tab Total \cr
   1985 \tab 920 \tab 2 \tab 39 \tab 22 \tab 542 \tab 4 \tab 5 \tab 1534 \cr
   2004 \tab 2472 \tab 8 \tab 30 \tab 17 \tab 234 \tab 5 \tab 46 \tab 2812 \cr
   Total \tab 3392 \tab 10 \tab 69 \tab 39 \tab 776 \tab 9 \tab 51 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race5}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}}
}
\concept{Social Networks}
\keyword{variable}
