% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_96.R
\name{race5}
\alias{race5}
\title{Race of person number 5}
\source{
General Social Survey https://gss.norc.org
}
\description{
race5
}
\details{
Question Is (NAME) Asian, Black, Hispanic, White or something else?  ASK FOR EACH NAME.
5. Race of Fifth Person
}
\section{Values}{

\itemize{
\item \verb{[1]} asian
\item \verb{[2]} black
\item \verb{[3]} hispanic
\item \verb{[4]} white
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/874/vshow}{GSS Data Explorer page for \code{race5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab asian \tab black \tab hispanic \tab white \tab other \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 1 \tab 24 \tab 12 \tab 336 \tab 4 \tab 4 \tab 1534 \cr
   2004 \tab 2600 \tab 2 \tab 11 \tab 7 \tab 143 \tab 3 \tab 46 \tab 2812 \cr
   Total \tab 3753 \tab 3 \tab 35 \tab 19 \tab 479 \tab 7 \tab 50 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}}
}
\concept{Social Networks}
\keyword{variable}
