% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_625.R
\name{raceacs12}
\alias{raceacs12}
\title{Guamanian or chamorro}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs12
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Guamanian or Chamorro
As of 2024, RACEACS12 has been recoded into RACEACS14 as "Other Pacific Islander."
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7461/vshow}{GSS Data Explorer page for \code{raceacs12}}.
}

\seealso{
Other Core: 
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs4}},
\code{\link{raceacs6}},
\code{\link{raceacs8}}
}
\concept{Core}
\keyword{variable}
