% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_626.R
\name{raceacs13}
\alias{raceacs13}
\title{Samoan}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs13
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Samoan
As of 2024, RACEACS13 has been recoded into RACEACS14 as "Other Pacific Islander."
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7462/vshow}{GSS Data Explorer page for \code{raceacs13}}.
}

\seealso{
Other Core: 
\code{\link{raceacs14}},
\code{\link{raceacs15}}
}
\concept{Core}
\keyword{variable}
