% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_626.R
\name{raceacs14}
\alias{raceacs14}
\title{Other pacific islander}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs14
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Other Pacific Islander
As of 2024, RACEACS14 has been recoded to include RACEACS11 (Native Hawaiian), RACEACS12 (Guamanian or Chamorro), and RACEACS13 (Samoan). This has been retroactively applied to all previous years of RACEACS14. The original RACEACS14 containing data from 2021 to 2022 has been renamed RACEACS14_2122.
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7463/vshow}{GSS Data Explorer page for \code{raceacs14}}.
}

\seealso{
Other Core: 
\code{\link{raceacs13}},
\code{\link{raceacs15}}

Other Race: 
\code{\link{raceacs13}},
\code{\link{raceacs16}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
