% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_694.R
\name{raceacs14_2122}
\alias{raceacs14_2122}
\title{Other pacific islander}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs14_2122
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Other Pacific Islander
As of 2024, RACEACS14 has been recoded to include RACEACS11 (Native Hawaiian), RACEACS12 (Guamanian or Chamorro), and RACEACS13 (Samoan). This has been retroactively applied to all previous years of RACEACS14. The original RACEACS14 containing data from 2021 to 2022 has been renamed RACEACS14_2122.
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8221/vshow}{GSS Data Explorer page for \code{raceacs14_2122}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab no \tab yes \tab don't know \tab no answer \tab refused \tab Total \cr
   2021 \tab - \tab 4014 \tab 3 \tab 6 \tab 9 \tab - \tab 4032 \cr
   2022 \tab - \tab 3496 \tab 7 \tab 8 \tab 18 \tab 15 \tab 3544 \cr
   Total \tab 0 \tab 7510 \tab 10 \tab 14 \tab 27 \tab 15 \tab 7576 \cr
}
}

\seealso{
Other Core: 
\code{\link{hompop_7222}},
\code{\link{hompop_exp_22}},
\code{\link{reg16_7222}},
\code{\link{region_7222}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
