% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_624.R
\name{raceacs2}
\alias{raceacs2}
\title{Black}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs2
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Black or African American
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7451/vshow}{GSS Data Explorer page for \code{raceacs2}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab no \tab yes \tab don't know \tab no answer \tab refused \tab Total \cr
   2021 \tab - \tab 3522 \tab 495 \tab 6 \tab 9 \tab - \tab 4032 \cr
   2022 \tab - \tab 2913 \tab 590 \tab 8 \tab 18 \tab 15 \tab 3544 \cr
   2024 \tab - \tab 2666 \tab 592 \tab 18 \tab 24 \tab 9 \tab 3309 \cr
   Total \tab 0 \tab 9101 \tab 1677 \tab 32 \tab 51 \tab 24 \tab 10885 \cr
}
}

\seealso{
Other Core: 
\code{\link{raceacs1}}
}
\concept{Core}
\keyword{variable}
