% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_625.R
\name{raceacs6}
\alias{raceacs6}
\title{Filipino}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs6
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Filipino
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7455/vshow}{GSS Data Explorer page for \code{raceacs6}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab refused \tab Total \cr
   2021 \tab - \tab 3997 \tab 20 \tab 6 \tab 9 \tab - \tab 4032 \cr
   2022 \tab - \tab 3475 \tab 28 \tab 8 \tab 18 \tab 15 \tab 3544 \cr
   Total \tab 0 \tab 7472 \tab 48 \tab 14 \tab 27 \tab 15 \tab 7576 \cr
}
}

\seealso{
Other Core: 
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs2}},
\code{\link{raceacs4}},
\code{\link{raceacs8}}
}
\concept{Core}
\keyword{variable}
