% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_625.R
\name{raceacs7}
\alias{raceacs7}
\title{Japanese}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs7
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Japanese
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7456/vshow}{GSS Data Explorer page for \code{raceacs7}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab no \tab yes \tab don't know \tab no answer \tab refused \tab Total \cr
   2021 \tab - \tab 3994 \tab 23 \tab 6 \tab 9 \tab - \tab 4032 \cr
   2022 \tab - \tab 3491 \tab 12 \tab 8 \tab 18 \tab 15 \tab 3544 \cr
   2024 \tab - \tab 3243 \tab 15 \tab 18 \tab 24 \tab 9 \tab 3309 \cr
   Total \tab 0 \tab 10728 \tab 50 \tab 32 \tab 51 \tab 24 \tab 10885 \cr
}
}

\seealso{
Other Race: 
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs9}}
}
\concept{Race}
\keyword{variable}
