% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{racecen1}
\alias{racecen1}
\title{What is r's race 1st mention}
\source{
General Social Survey https://gss.norc.org
}
\description{
racecen1
}
\details{
Question What is your race?  Indicate one or more races that you consider yourself to be. A. First mention
As of 2024, RACECEN1 has been recoded to combine Native Hawaiian, Guamanian or Chamorro, and Samoan into the "Other Pacific Islander" category. This has been retroactively applied to all previous years of RACECEN1. The original RACECEN1 containing data from 2000 to 2022 has been renamed RACECEN1_0022.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black or african american
\item \code{3} american indian or alaska native
\item \code{4} asian
\item \code{5} native hawaiian or other pacific islander
\item \code{6} some other race
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5277/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab american indian or alaska native \tab asian \tab black or african american \tab iap \tab native hawaiian or other pacific islander \tab some other race \tab white \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 16 \tab 32 \tab 201 \tab 1419 \tab 2 \tab 54 \tab 1093 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 33 \tab 65 \tab 402 \tab - \tab 3 \tab 84 \tab 2175 \tab 2 \tab 1 \tab - \tab 2765 \cr
   2004 \tab - \tab 26 \tab 96 \tab 374 \tab - \tab 5 \tab 84 \tab 2222 \tab 2 \tab 3 \tab - \tab 2812 \cr
   2006 \tab - \tab 41 \tab 138 \tab 630 \tab - \tab 13 \tab 209 \tab 3470 \tab 5 \tab 4 \tab - \tab 4510 \cr
   2008 \tab - \tab 27 \tab 61 \tab 280 \tab - \tab 9 \tab 90 \tab 1545 \tab 8 \tab 3 \tab - \tab 2023 \cr
   2010 \tab - \tab 20 \tab 64 \tab 304 \tab - \tab 4 \tab 101 \tab 1539 \tab 4 \tab 8 \tab - \tab 2044 \cr
   2012 \tab - \tab 29 \tab 65 \tab 297 \tab - \tab 3 \tab 99 \tab 1471 \tab 6 \tab 4 \tab - \tab 1974 \cr
   2014 \tab - \tab 42 \tab 72 \tab 385 \tab - \tab 5 \tab 134 \tab 1878 \tab 9 \tab 13 \tab - \tab 2538 \cr
   2016 \tab - \tab 58 \tab 90 \tab 487 \tab - \tab 4 \tab 116 \tab 2088 \tab 9 \tab 15 \tab - \tab 2867 \cr
   2018 \tab - \tab 36 \tab 73 \tab 382 \tab - \tab 4 \tab 154 \tab 1679 \tab 9 \tab 11 \tab - \tab 2348 \cr
   2021 \tab - \tab 44 \tab 160 \tab 463 \tab 6 \tab 9 \tab 192 \tab 3110 \tab 39 \tab 9 \tab - \tab 4032 \cr
   2022 \tab - \tab 63 \tab 131 \tab 565 \tab 25 \tab 6 \tab 212 \tab 2514 \tab 9 \tab 19 \tab - \tab 3544 \cr
   2024 \tab - \tab 47 \tab 149 \tab 572 \tab 27 \tab 11 \tab 185 \tab 2280 \tab 8 \tab 26 \tab 4 \tab 3309 \cr
   Total \tab 38116 \tab 482 \tab 1196 \tab 5342 \tab 1477 \tab 78 \tab 1714 \tab 27064 \tab 110 \tab 116 \tab 4 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}

Other Race: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\concept{Race}
\keyword{variable}
