% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{racecen2}
\alias{racecen2}
\title{What is r's race 2nd mention}
\source{
General Social Survey https://gss.norc.org
}
\description{
racecen2
}
\details{
Question What is your race?  Indicate one or more races that you consider yourself to be. A. Second mention
As of 2024, RACECEN2 has been recoded to combine Native Hawaiian, Guamanian or Chamorro, and Samoan into the "Other Pacific Islander" category. This has been retroactively applied to all previous years of RACECEN2. The original RACECEN2 containing data from 2000 to 2022 has been renamed RACECEN2_0022.
}
\section{Values}{

\itemize{
\item \verb{[1]} white
\item \verb{[2]} black or african american
\item \verb{[3]} american indian or alaska native
\item \verb{[4]} asian indian
\item \verb{[5]} chinese
\item \verb{[6]} filipino
\item \verb{[7]} japanese
\item \verb{[8]} korean
\item \verb{[9]} vietnamese
\item \verb{[10]} other asian
\item \verb{[14]} other pacific islander
\item \verb{[15]} some other race
\item \verb{[16]} hispanic
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5278/vshow}{GSS Data Explorer page for \code{racecen2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab not available in this year \tab white \tab black or african american \tab american indian or alaska native \tab chinese \tab filipino \tab japanese \tab other pacific islander \tab some other race \tab hispanic \tab iap \tab vietnamese \tab other asian \tab don't know \tab no answer \tab korean \tab asian indian \tab skipped on web \tab Total \cr
   2000 \tab - \tab 8 \tab 2 \tab 40 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 15 \tab 2745 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 28 \tab 10 \tab 94 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab 10 \tab 2601 \tab 2 \tab 6 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 28 \tab 7 \tab 95 \tab 2 \tab 2 \tab 4 \tab 1 \tab 5 \tab 11 \tab 2650 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 41 \tab 13 \tab 135 \tab 7 \tab 1 \tab 1 \tab 8 \tab 8 \tab 40 \tab 4244 \tab - \tab 5 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 4510 \cr
   2008 \tab - \tab 14 \tab 10 \tab 71 \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 13 \tab 1902 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2023 \cr
   2010 \tab - \tab 21 \tab 18 \tab 77 \tab 1 \tab 6 \tab 2 \tab 6 \tab 2 \tab 5 \tab 1890 \tab - \tab 1 \tab - \tab 9 \tab 3 \tab 3 \tab - \tab 2044 \cr
   2012 \tab - \tab 22 \tab 17 \tab 70 \tab 2 \tab 5 \tab 2 \tab 3 \tab 2 \tab 12 \tab 1830 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 3 \tab - \tab 1974 \cr
   2014 \tab - \tab 34 \tab 17 \tab 86 \tab 1 \tab 5 \tab 3 \tab 2 \tab 4 \tab 18 \tab 2346 \tab - \tab 3 \tab 1 \tab 13 \tab 2 \tab 3 \tab - \tab 2538 \cr
   2016 \tab - \tab 35 \tab 24 \tab 140 \tab 2 \tab 3 \tab 3 \tab 6 \tab 4 \tab 8 \tab 2617 \tab 1 \tab 4 \tab 1 \tab 15 \tab 3 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 25 \tab 22 \tab 103 \tab 5 \tab 3 \tab 3 \tab 4 \tab 15 \tab 24 \tab 2122 \tab - \tab 4 \tab 2 \tab 13 \tab 2 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab 51 \tab 10 \tab 62 \tab 7 \tab 3 \tab 7 \tab 4 \tab 3 \tab 8 \tab 3858 \tab 1 \tab 3 \tab 10 \tab - \tab 2 \tab 3 \tab - \tab 4032 \cr
   2022 \tab - \tab 52 \tab 16 \tab 72 \tab 7 \tab 3 \tab 3 \tab 5 \tab 7 \tab 8 \tab 3362 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab - \tab 3544 \cr
   2024 \tab - \tab 41 \tab 11 \tab 50 \tab 6 \tab 9 \tab 3 \tab 3 \tab 4 \tab 8 \tab 3124 \tab 3 \tab 1 \tab 9 \tab 26 \tab 3 \tab 4 \tab 4 \tab 3309 \cr
   Total \tab 0 \tab 400 \tab 177 \tab 1095 \tab 49 \tab 44 \tab 34 \tab 48 \tab 60 \tab 180 \tab 35291 \tab 10 \tab 33 \tab 29 \tab 86 \tab 18 \tab 25 \tab 4 \tab 37583 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethadapt}},
\code{\link{ethclose}},
\code{\link{ethdiff}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethnic}},
\code{\link{ethnofit}},
\code{\link{ethnum}},
\code{\link{ethorgs}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{fucitzn}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hspgovt}},
\code{\link{meltpot}},
\code{\link{natborn}},
\code{\link{owneth}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{schleth}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{teacheth}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{whoteach}},
\code{\link{whtgovt}}

Other Race: 
\code{\link{admirblk}},
\code{\link{allowmar}},
\code{\link{asngovt}},
\code{\link{blkgains}},
\code{\link{blkgovt}},
\code{\link{blkinflu}},
\code{\link{blksimp}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{bwfriend}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{color}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{dsndntpay}},
\code{\link{ethchng}},
\code{\link{ethhist}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdmar}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{hseth}},
\code{\link{hspgovt}},
\code{\link{hswht}},
\code{\link{leastcom}},
\code{\link{meltpot}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{numbwfrd}},
\code{\link{oprace}},
\code{\link{owneth}},
\code{\link{perla}},
\code{\link{perla_self}},
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs16}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs7}},
\code{\link{raceacs9}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmeet}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpromo}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{ractrust}},
\code{\link{racwork}},
\code{\link{ratetone}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{trtmyrac}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wrkwayup}}
}
\concept{Ethnicity}
\concept{Race}
\keyword{variable}
