% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_687.R
\name{racecen2_0022}
\alias{racecen2_0022}
\title{What is r's race, second mention}
\source{
General Social Survey https://gss.norc.org
}
\description{
racecen2_0022
}
\details{
Question What is your race?  Indicate one or more races that you consider yourself to be.
B. Second mention

As of 2024, RACECEN2 has been recoded to coarsen certain sensitive categories. This has been retroactively applied to all previous years of RACECEN2. The original RACECEN2 containing data from 2000 to 2022 has been renamed RACECEN2_0022.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black or african american
\item \code{3} american indian or alaska native
\item \code{4} asian indian
\item \code{5} chinese
\item \code{6} filipino
\item \code{7} japanese
\item \code{8} korean
\item \code{9} vietnamese
\item \code{10} other asian
\item \code{11} native hawaiian
\item \code{12} guamanian or chamorro
\item \code{13} samoan
\item \code{14} other pacific islander
\item \code{15} some other race
\item \code{16} hispanic
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8161/vshow}{GSS Data Explorer page for \code{racecen2_0022}}.
}

\seealso{
Other Ethnicity: 
\code{\link{racecen1_0022}},
\code{\link{racecen3_0022}}
}
\concept{Ethnicity}
\keyword{variable}
