% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{racerank2}
\alias{racerank2}
\title{Second race selected}
\source{
General Social Survey https://gss.norc.org
}
\description{
racerank2
}
\details{
Question After your primary racial identity, which of these races do you identify most with?
}
\section{Values}{

\itemize{
\item \verb{[1]} white
\item \verb{[2]} black or african-american
\item \verb{[3]} american indian or alaska native
\item \verb{[4]} asian indian
\item \verb{[5]} chinese
\item \verb{[6]} filipino
\item \verb{[7]} japanese
\item \verb{[8]} korean
\item \verb{[9]} vietnamese
\item \verb{[10]} other asian
\item \verb{[11]} native hawaiian
\item \verb{[12]} guamanian or chamorro
\item \verb{[13]} samoan
\item \verb{[14]} other pacific islander
\item \verb{[15]} some other race
\item \verb{[16]} hispanic
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7752/vshow}{GSS Data Explorer page for \code{racerank2}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab not available in this year \tab white \tab black or african-american \tab american indian or alaska native \tab asian indian \tab chinese \tab filipino \tab japanese \tab vietnamese \tab other asian \tab native hawaiian \tab samoan \tab other pacific islander \tab some other race \tab hispanic \tab don't know \tab iap \tab no answer \tab skipped on web \tab korean \tab guamanian or chamorro \tab Total \cr
   2022 \tab - \tab 52 \tab 16 \tab 72 \tab 3 \tab 7 \tab 3 \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 7 \tab 8 \tab 9 \tab 3332 \tab 20 \tab 2 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab 41 \tab 11 \tab 50 \tab 4 \tab 6 \tab 9 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 8 \tab 9 \tab 3124 \tab 26 \tab 4 \tab 3 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 93 \tab 27 \tab 122 \tab 7 \tab 13 \tab 12 \tab 6 \tab 5 \tab 4 \tab 4 \tab 2 \tab 1 \tab 11 \tab 16 \tab 18 \tab 6456 \tab 46 \tab 6 \tab 3 \tab 1 \tab 6853 \cr
}
}

\keyword{variable}
