% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{racerank3}
\alias{racerank3}
\title{Third race selected}
\source{
General Social Survey https://gss.norc.org
}
\description{
racerank3
}
\details{
Question After your primary and secondary racial identity, which of these races do you identify most with?
}
\section{Values}{

\itemize{
\item \verb{[1]} white
\item \verb{[2]} black or african-american
\item \verb{[3]} american indian or alaska native
\item \verb{[4]} asian indian
\item \verb{[5]} chinese
\item \verb{[6]} filipino
\item \verb{[7]} japanese
\item \verb{[8]} korean
\item \verb{[9]} vietnamese
\item \verb{[10]} other asian
\item \verb{[11]} native hawaiian
\item \verb{[12]} guamanian or chamorro
\item \verb{[13]} samoan
\item \verb{[14]} other pacific islander
\item \verb{[15]} some other race
\item \verb{[16]} hispanic
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7753/vshow}{GSS Data Explorer page for \code{racerank3}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab not available in this year \tab white \tab black or african-american \tab american indian or alaska native \tab asian indian \tab other pacific islander \tab some other race \tab hispanic \tab don't know \tab iap \tab no answer \tab skipped on web \tab chinese \tab other asian \tab native hawaiian \tab Total \cr
   2022 \tab - \tab 9 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 9 \tab 3494 \tab 20 \tab 2 \tab - \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab 12 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 9 \tab 3249 \tab 26 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 21 \tab 2 \tab 8 \tab 2 \tab 2 \tab 1 \tab 1 \tab 18 \tab 6743 \tab 46 \tab 6 \tab 1 \tab 1 \tab 1 \tab 6853 \cr
}
}

\keyword{variable}
