% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{racesurv17}
\alias{racesurv17}
\title{How much, if at all, do you think the legacy of slavery affects the position of}
\source{
General Social Survey https://gss.norc.org
}
\description{
racesurv17
}
\details{
Question How much, if at all, do you think the legacy of slavery affects the position of black people in American society today?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} a fair amount
\item \verb{[3]} not much
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7577/vshow}{GSS Data Explorer page for \code{racesurv17}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a great deal \tab a fair amount \tab not much \tab not at all \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1315 \tab 802 \tab 704 \tab 474 \tab 360 \tab 7 \tab 336 \tab 34 \tab 4032 \cr
   Total \tab 1315 \tab 802 \tab 704 \tab 474 \tab 360 \tab 7 \tab 336 \tab 34 \tab 4032 \cr
}
}

\keyword{variable}
