% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{racethas}
\alias{racethas}
\title{Points assigned to asian american}
\source{
General Social Survey https://gss.norc.org
}
\description{
racethas
}
\details{
Question d.  Asian American?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6354/vshow}{GSS Data Explorer page for \code{racethas}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1888 \tab 931 \tab 3 \tab 14 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 5 \tab 15 \tab 2867 \cr
   Total \tab 1888 \tab 931 \tab 3 \tab 14 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 5 \tab 15 \tab 2867 \cr
}
}

\seealso{
Other Racial/Ethnic Identity: 
\code{\link{racethbl}},
\code{\link{racethhi}},
\code{\link{racethna}},
\code{\link{racethot}},
\code{\link{racethwh}}
}
\concept{Racial/Ethnic Identity}
\keyword{variable}
