% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{racethmn}
\alias{racethmn}
\title{Specify other in racethot}
\source{
General Social Survey https://gss.norc.org
}
\description{
racethmn
}
\details{
Question In describing our background, we often ask someone which racial or ethnic groups best describes them. Imagine if we used instead a 10 point system where points are allocated to whichever racial or ethnic groups the respondent thinks accurately describes themselves. For example, if you think of yourself as just Latino, you would allocate all 10 points to Latino. Or if you think of yourself as mixed race, but more African American than white, you might allocate 6 or 7 points to black and 3 or 4 points to white. Thinking of your background in racial and ethnic terms ... How would you describe your race and ethnicity using this 10-point system using the following choices: G. Specified other ethnicities
}
\section{Values}{

\itemize{
\item \verb{[1]} european
\item \verb{[2]} hispanic
\item \verb{[3]} african
\item \verb{[4]} east asian
\item \verb{[5]} native american
\item \verb{[6]} south asian
\item \verb{[7]} asian, not south or east
\item \verb{[8]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6649/vshow}{GSS Data Explorer page for \code{racethmn}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab european \tab hispanic \tab african \tab east asian \tab south asian \tab asian, not south or east \tab other \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1888 \tab 22 \tab 2 \tab 1 \tab 6 \tab 5 \tab 2 \tab 12 \tab 2 \tab 927 \tab 2867 \cr
   Total \tab 1888 \tab 22 \tab 2 \tab 1 \tab 6 \tab 5 \tab 2 \tab 12 \tab 2 \tab 927 \tab 2867 \cr
}
}

\keyword{variable}
