% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{racethwh}
\alias{racethwh}
\title{Points assigned to white}
\source{
General Social Survey https://gss.norc.org
}
\description{
racethwh
}
\details{
Question a.  White?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6351/vshow}{GSS Data Explorer page for \code{racethwh}}.
}

\seealso{
Other Racial/Ethnic Identity: 
\code{\link{racethas}},
\code{\link{racethbl}},
\code{\link{racethhi}},
\code{\link{racethna}},
\code{\link{racethot}}
}
\concept{Racial/Ethnic Identity}
\keyword{variable}
