% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racfew}
\alias{racfew}
\title{Children to school with opposite race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racfew
}
\details{
Question A. Would you yourself have any objection to sending your children to a school where a few of the children are
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, object
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1988, 1989, 1991, 1993, 1994, 1996 \cr
   A/B/C \tab 1990 \cr
   Full \tab 1972, 1974, 1975, 1977, 1978, 1982, 1983, 1985, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/395/vshow}{GSS Data Explorer page for \code{racfew}}.

Counts by year:\tabular{lllllll}{
   year \tab yes, object \tab no \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 96 \tab 1242 \tab 13 \tab 261 \tab 1 \tab 1613 \cr
   1974 \tab 67 \tab 1221 \tab 7 \tab 173 \tab 16 \tab 1484 \cr
   1975 \tab 90 \tab 1214 \tab 22 \tab 163 \tab 1 \tab 1490 \cr
   1977 \tab 98 \tab 1247 \tab 5 \tab 176 \tab 4 \tab 1530 \cr
   1978 \tab 70 \tab 1441 \tab 15 \tab - \tab 6 \tab 1532 \cr
   1982 \tab 104 \tab 1735 \tab 17 \tab - \tab 4 \tab 1860 \cr
   1983 \tab 78 \tab 1503 \tab 15 \tab - \tab 3 \tab 1599 \cr
   1985 \tab 64 \tab 1454 \tab 8 \tab - \tab 8 \tab 1534 \cr
   1986 \tab 70 \tab 1386 \tab 11 \tab - \tab 3 \tab 1470 \cr
   1988 \tab 41 \tab 932 \tab 13 \tab 493 \tab 2 \tab 1481 \cr
   1989 \tab 42 \tab 942 \tab 18 \tab 531 \tab 4 \tab 1537 \cr
   1990 \tab 42 \tab 1314 \tab 11 \tab - \tab 5 \tab 1372 \cr
   1991 \tab 33 \tab 973 \tab 10 \tab 493 \tab 8 \tab 1517 \cr
   1993 \tab 47 \tab 1026 \tab 5 \tab 526 \tab 2 \tab 1606 \cr
   1994 \tab 71 \tab 1875 \tab 18 \tab 1015 \tab 13 \tab 2992 \cr
   1996 \tab 29 \tab 951 \tab 5 \tab 1913 \tab 6 \tab 2904 \cr
   Total \tab 1042 \tab 20456 \tab 193 \tab 5744 \tab 86 \tab 27521 \cr
}
}

\seealso{
Other Core: 
\code{\link{racclos}},
\code{\link{racdis}},
\code{\link{rachome}},
\code{\link{racinteg}},
\code{\link{raclive}},
\code{\link{racobjct}},
\code{\link{racopen}},
\code{\link{racschol}},
\code{\link{racseg}}
}
\concept{Core}
\keyword{variable}
