% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{rachisch}
\alias{rachisch}
\title{How integrated was r's high school}
\source{
General Social Survey https://gss.norc.org
}
\description{
rachisch
}
\details{
Question Was the high school you attended all black, integrated--but mostly black, or integrated--but mostly white?
}
\section{Values}{

\itemize{
\item \verb{[0]} didnt go to hs
\item \verb{[1]} all black
\item \verb{[2]} mostly black
\item \verb{[3]} mostly white
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/405/vshow}{GSS Data Explorer page for \code{rachisch}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab didnt go to hs \tab all black \tab mostly black \tab mostly white \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 33 \tab 235 \tab 120 \tab 112 \tab 10 \tab 1860 \cr
   Total \tab 1350 \tab 33 \tab 235 \tab 120 \tab 112 \tab 10 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\keyword{variable}
