% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{rachome}
\alias{rachome}
\title{Opp. race home for dinner recently}
\source{
General Social Survey https://gss.norc.org
}
\description{
rachome
}
\details{
Question During the last few years, has anyone in your family brought a friend who was a (Negro/Black/African-American) home for dinner?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/393/vshow}{GSS Data Explorer page for \code{rachome}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1973 \tab 183 \tab 261 \tab 1049 \tab 11 \tab - \tab 1504 \cr
   1974 \tab 173 \tab 294 \tab 1006 \tab 6 \tab 5 \tab 1484 \cr
   1976 \tab 129 \tab 311 \tab 1049 \tab 8 \tab 2 \tab 1499 \cr
   1977 \tab 176 \tab 313 \tab 1035 \tab 5 \tab 1 \tab 1530 \cr
   1980 \tab - \tab 420 \tab 1038 \tab 2 \tab 8 \tab 1468 \cr
   1982 \tab - \tab 588 \tab 1251 \tab 10 \tab 11 \tab 1860 \cr
   1984 \tab - \tab 453 \tab 1010 \tab 8 \tab 2 \tab 1473 \cr
   1985 \tab - \tab 446 \tab 1075 \tab 9 \tab 4 \tab 1534 \cr
   1987 \tab - \tab 620 \tab 1185 \tab 8 \tab 6 \tab 1819 \cr
   1988 \tab 504 \tab 312 \tab 655 \tab 5 \tab 5 \tab 1481 \cr
   1989 \tab 504 \tab 304 \tab 719 \tab 5 \tab 5 \tab 1537 \cr
   1990 \tab 455 \tab 264 \tab 642 \tab 4 \tab 7 \tab 1372 \cr
   1991 \tab 524 \tab 300 \tab 676 \tab 10 \tab 7 \tab 1517 \cr
   1993 \tab 531 \tab 392 \tab 673 \tab 6 \tab 4 \tab 1606 \cr
   1994 \tab 996 \tab 730 \tab 1255 \tab 5 \tab 6 \tab 2992 \cr
   1996 \tab 1953 \tab 386 \tab 553 \tab 8 \tab 4 \tab 2904 \cr
   2006 \tab 2992 \tab 619 \tab 890 \tab 4 \tab 5 \tab 4510 \cr
   Total \tab 9120 \tab 7013 \tab 15761 \tab 114 \tab 82 \tab 32090 \cr
}
}

\seealso{
Other Core: 
\code{\link{racclos}},
\code{\link{racdis}},
\code{\link{racfew}},
\code{\link{racinteg}},
\code{\link{raclive}},
\code{\link{racobjct}},
\code{\link{racopen}},
\code{\link{racschol}},
\code{\link{racseg}}
}
\concept{Core}
\keyword{variable}
