% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{racjob}
\alias{racjob}
\title{Equal opportunities on jobs}
\source{
General Social Survey https://gss.norc.org
}
\description{
racjob
}
\details{
Question Do you think Negroes should have as good a chance as white people to get any kind of job, or do you think white people should have the first chance at any kind of job?
}
\section{Values}{

\itemize{
\item \code{1} as good a chance
\item \code{2} whites first
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/401/vshow}{GSS Data Explorer page for \code{racjob}}.

Counts by year:\tabular{lllllll}{
   year \tab as good a chance \tab whites first \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 1289 \tab 43 \tab 16 \tab 261 \tab 4 \tab 1613 \cr
   Total \tab 1289 \tab 43 \tab 16 \tab 261 \tab 4 \tab 1613 \cr
}
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\keyword{variable}
