% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{raclive}
\alias{raclive}
\title{Any opp. race in neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
raclive
}
\details{
Question Are there any ("whites" for black respondents, "blacks" for non-black respondents) living in this neighborhood now?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/388/vshow}{GSS Data Explorer page for \code{raclive}}.

Counts by year:\tabular{llllllll}{
   year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 386 \tab 908 \tab 57 \tab 261 \tab 1 \tab - \tab 1613 \cr
   1973 \tab 524 \tab 745 \tab 51 \tab 183 \tab 1 \tab - \tab 1504 \cr
   1974 \tab 554 \tab 711 \tab 45 \tab 173 \tab 1 \tab - \tab 1484 \cr
   1975 \tab 436 \tab 834 \tab 56 \tab 163 \tab 1 \tab - \tab 1490 \cr
   1976 \tab 580 \tab 737 \tab 50 \tab 129 \tab 3 \tab - \tab 1499 \cr
   1977 \tab 532 \tab 777 \tab 41 \tab 176 \tab 4 \tab - \tab 1530 \cr
   1978 \tab 745 \tab 731 \tab 54 \tab - \tab 2 \tab - \tab 1532 \cr
   1980 \tab 675 \tab 737 \tab 54 \tab - \tab 2 \tab - \tab 1468 \cr
   1982 \tab 938 \tab 852 \tab 66 \tab - \tab 4 \tab - \tab 1860 \cr
   1983 \tab 751 \tab 796 \tab 49 \tab - \tab 3 \tab - \tab 1599 \cr
   1984 \tab 746 \tab 660 \tab 59 \tab - \tab 8 \tab - \tab 1473 \cr
   1985 \tab 734 \tab 726 \tab 67 \tab - \tab 7 \tab - \tab 1534 \cr
   1986 \tab 701 \tab 727 \tab 40 \tab - \tab 2 \tab - \tab 1470 \cr
   1987 \tab 981 \tab 772 \tab 60 \tab - \tab 6 \tab - \tab 1819 \cr
   1988 \tab 790 \tab 619 \tab 67 \tab - \tab 5 \tab - \tab 1481 \cr
   1989 \tab 786 \tab 666 \tab 75 \tab - \tab 10 \tab - \tab 1537 \cr
   1990 \tab 743 \tab 550 \tab 75 \tab - \tab 4 \tab - \tab 1372 \cr
   1991 \tab 788 \tab 648 \tab 66 \tab - \tab 15 \tab - \tab 1517 \cr
   1993 \tab 955 \tab 562 \tab 79 \tab - \tab 10 \tab - \tab 1606 \cr
   1994 \tab 1820 \tab 1014 \tab 143 \tab - \tab 15 \tab - \tab 2992 \cr
   1996 \tab 1771 \tab 949 \tab 158 \tab - \tab 26 \tab - \tab 2904 \cr
   1998 \tab 1749 \tab 887 \tab 141 \tab - \tab 55 \tab - \tab 2832 \cr
   2000 \tab 1791 \tab 831 \tab 150 \tab - \tab 45 \tab - \tab 2817 \cr
   2002 \tab 933 \tab 365 \tab 49 \tab 1395 \tab 23 \tab - \tab 2765 \cr
   2004 \tab 853 \tab 431 \tab 55 \tab 1472 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 1973 \tab 927 \tab 91 \tab 1518 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1366 \tab 580 \tab 77 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1361 \tab 596 \tab 85 \tab - \tab 2 \tab - \tab 2044 \cr
   2012 \tab 1374 \tab 517 \tab 82 \tab - \tab 1 \tab - \tab 1974 \cr
   2014 \tab 1831 \tab 587 \tab 118 \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2095 \tab 648 \tab 122 \tab - \tab 2 \tab - \tab 2867 \cr
   2018 \tab 1738 \tab 477 \tab 127 \tab - \tab 6 \tab - \tab 2348 \cr
   2021 \tab 2955 \tab 632 \tab 440 \tab - \tab - \tab 5 \tab 4032 \cr
   2022 \tab 2620 \tab 665 \tab 257 \tab - \tab - \tab 2 \tab 3544 \cr
   Total \tab 39575 \tab 23864 \tab 3206 \tab 5470 \tab 268 \tab 7 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{racclos}},
\code{\link{racdis}},
\code{\link{racfew}},
\code{\link{rachome}},
\code{\link{racinteg}},
\code{\link{racobjct}},
\code{\link{racopen}},
\code{\link{racschol}},
\code{\link{racseg}}
}
\concept{Core}
\keyword{variable}
