% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_539.R
\name{racmeet}
\alias{racmeet}
\title{Allowed to hold pub meeting for racist}
\source{
General Social Survey https://gss.norc.org
}
\description{
racmeet
}
\details{
Question Should people prejudiced against any racial or ethnic group be allowed to hold public meetings?
}
\section{Values}{

\itemize{
\item \verb{[1]} should definitely be allowed
\item \verb{[2]} should probably be allowed
\item \verb{[3]} should probably not be allowed
\item \verb{[4]} should definitely not be allowed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4919/vshow}{GSS Data Explorer page for \code{racmeet}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab should definitely be allowed \tab should probably be allowed \tab should probably not be allowed \tab should definitely not be allowed \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   2004 \tab 1340 \tab 241 \tab 331 \tab 308 \tab 574 \tab 11 \tab 7 \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 217 \tab 263 \tab 275 \tab 442 \tab 65 \tab 2 \tab - \tab 2538 \cr
   Total \tab 2614 \tab 458 \tab 594 \tab 583 \tab 1016 \tab 76 \tab 9 \tab 0 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{buypol}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{citviews}},
\code{\link{cntctgov}},
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{eqtreat}},
\code{\link{fixmistk}},
\code{\link{govdook}},
\code{\link{govngos}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{gvtrghts}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{hlthall}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{leftrght}},
\code{\link{leftrght1}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ntcitvte}},
\code{\link{obeylaws}},
\code{\link{oppsegov}},
\code{\link{othreasn}},
\code{\link{owndecns}},
\code{\link{paytaxes}},
\code{\link{polactve}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polnews}},
\code{\link{polopts}},
\code{\link{poltalk}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{relmeet}},
\code{\link{revmeet}},
\code{\link{rghtsmin}},
\code{\link{servepeo}},
\code{\link{servpub}},
\code{\link{signdpet}},
\code{\link{solok}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usmca1}},
\code{\link{usmca2}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Race: 
\code{\link{admirblk}},
\code{\link{allowmar}},
\code{\link{asngovt}},
\code{\link{blkgains}},
\code{\link{blkgovt}},
\code{\link{blkinflu}},
\code{\link{blksimp}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{bwfriend}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{color}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{dsndntpay}},
\code{\link{ethchng}},
\code{\link{ethhist}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdmar}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{hseth}},
\code{\link{hspgovt}},
\code{\link{hswht}},
\code{\link{leastcom}},
\code{\link{meltpot}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{numbwfrd}},
\code{\link{oprace}},
\code{\link{owneth}},
\code{\link{perla}},
\code{\link{perla_self}},
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs16}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs7}},
\code{\link{raceacs9}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpromo}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{ractrust}},
\code{\link{racwork}},
\code{\link{ratetone}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{trtmyrac}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wrkwayup}}
}
\concept{ISSP Citizenship}
\concept{Race}
\keyword{variable}
