% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racmix}
\alias{racmix}
\title{How many friends are white}
\source{
General Social Survey https://gss.norc.org
}
\description{
racmix
}
\details{
Question How many of your close friends are white?
}
\section{Values}{

\itemize{
\item \verb{[1]} none
\item \verb{[2]} one or two
\item \verb{[3]} three or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/406/vshow}{GSS Data Explorer page for \code{racmix}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab none \tab one or two \tab three or more \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 170 \tab 124 \tab 214 \tab 2 \tab 1860 \cr
   Total \tab 1350 \tab 170 \tab 124 \tab 214 \tab 2 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
