% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{racname}
\alias{racname}
\title{Preferred racial name}
\source{
General Social Survey https://gss.norc.org
}
\description{
racname
}
\details{
Question Which would you most like to be called, "Black", "Negro", "Colored", or "Afro- American", or does it make any difference?
}
\section{Values}{

\itemize{
\item \verb{[1]} black
\item \verb{[2]} negro
\item \verb{[3]} colored
\item \verb{[4]} afro-american
\item \verb{[8]} makes no difference
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/404/vshow}{GSS Data Explorer page for \code{racname}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab black \tab negro \tab colored \tab afro-american \tab makes no difference \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 264 \tab 29 \tab 23 \tab 29 \tab 158 \tab 7 \tab 1860 \cr
   Total \tab 1350 \tab 264 \tab 29 \tab 23 \tab 29 \tab 158 \tab 7 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racpres}}
}
\concept{Core}
\keyword{variable}
