% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racneigh}
\alias{racneigh}
\title{Neighborhood r would like to live in}
\source{
General Social Survey https://gss.norc.org
}
\description{
racneigh
}
\details{
Question If you could find the housing that you would want and like, would you rather live in a neighborhood that is all black; mostly black; half black, half white; or mostly white?
}
\section{Values}{

\itemize{
\item \verb{[1]} all black
\item \verb{[2]} mostly black
\item \verb{[3]} .5 black - .5 white
\item \verb{[4]} mostly white
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/407/vshow}{GSS Data Explorer page for \code{racneigh}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all black \tab mostly black \tab .5 black - .5 white \tab mostly white \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 67 \tab 83 \tab 282 \tab 26 \tab 36 \tab 16 \tab 1860 \cr
   Total \tab 1350 \tab 67 \tab 83 \tab 282 \tab 26 \tab 36 \tab 16 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
