% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racocc}
\alias{racocc}
\title{Equality in occupation}
\source{
General Social Survey https://gss.norc.org
}
\description{
racocc
}
\details{
Question Do you feel that a black person who has the same education and qualifications can get as good a job as a white person?  Would you say...
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} sometimes
\item \verb{[3]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/411/vshow}{GSS Data Explorer page for \code{racocc}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab almost always \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 66 \tab 271 \tab 165 \tab 3 \tab 5 \tab 1860 \cr
   Total \tab 1350 \tab 66 \tab 271 \tab 165 \tab 3 \tab 5 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
