% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{racopenv}
\alias{racopenv}
\title{Vote on open housing law (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
racopenv
}
\details{
Question Suppose there is a community-wide vote on the general housing issue. There are two possible laws to vote on: One law says that a homeowner can decide whom to sell their house to, even if that homeowner prefers not to sell to someone because of their race or color. The second law says that a homeowner cannot refuse to sell
to someone because of their race or color. Which law would you vote for?
}
\section{Values}{

\itemize{
\item \verb{[1]} first law
\item \verb{[2]} second law
\item \verb{[3]} neither (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
   A/C/- \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7588/vshow}{GSS Data Explorer page for \code{racopenv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab first law \tab second law \tab neither (phone mode only: volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1987 \tab 214 \tab 1521 \tab 302 \tab 1 \tab 1 \tab 6 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 5531 \tab 214 \tab 1521 \tab 302 \tab 1 \tab 1 \tab 6 \tab 7576 \cr
}
}

\keyword{variable}
