% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{racopeny}
\alias{racopeny}
\title{Vote on open housing law - gender neutral version}
\source{
General Social Survey https://gss.norc.org
}
\description{
racopeny
}
\details{
Question Suppose there is a community-wide vote on the general housing issue. There are two possible laws to vote on: One law says that a homeowner can decide whom to sell their house to, even if that homeowner prefers not to sell to (#racefill). The second law s
}
\section{Values}{

\itemize{
\item No applicable value labels.
\item 
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7754/vshow}{GSS Data Explorer page for \code{racopeny}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab first law \tab second law \tab neither \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2367 \tab 134 \tab 919 \tab 100 \tab 18 \tab 5 \tab 1 \tab 3544 \cr
   Total \tab 2367 \tab 134 \tab 919 \tab 100 \tab 18 \tab 5 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
