% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racparty}
\alias{racparty}
\title{Which party helps blacks more}
\source{
General Social Survey https://gss.norc.org
}
\description{
racparty
}
\details{
Question Which party, the Republican or the Democratic, do you feel will Do more to help blacks in the next few years, or do you think there isn't much difference between the two?
}
\section{Values}{

\itemize{
\item \verb{[1]} republican party
\item \verb{[2]} democratic party
\item \verb{[3]} not much difference
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/410/vshow}{GSS Data Explorer page for \code{racparty}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab republican party \tab democratic party \tab not much difference \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 8 \tab 336 \tab 142 \tab 17 \tab 7 \tab 1860 \cr
   Total \tab 1350 \tab 8 \tab 336 \tab 142 \tab 17 \tab 7 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
