% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{racpub}
\alias{racpub}
\title{Allow racists to publish books}
\source{
General Social Survey https://gss.norc.org
}
\description{
racpub
}
\details{
Question Second, consider people who believe that whites are racially superior to all other races.
Circle a number after each question to show if you think such people should be allowed to . . .
C. Publish books expressing their views.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3993/vshow}{GSS Data Explorer page for \code{racpub}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 181 \tab 189 \tab 95 \tab 162 \tab 41 \tab 9 \tab 1534 \cr
   1990 \tab 155 \tab 316 \tab 405 \tab 203 \tab 191 \tab 73 \tab 29 \tab 1372 \cr
   Total \tab 1012 \tab 497 \tab 594 \tab 298 \tab 353 \tab 114 \tab 38 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
