% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racquit}
\alias{racquit}
\title{Quit if couldn't integrate club}
\source{
General Social Survey https://gss.norc.org
}
\description{
racquit
}
\details{
Question If you and your friends belonged to a social club that would not let (Whites/Negroes/Blacks) join, would you try to change the rules so that (Whites/(Negroes/Blacks/African-Americans) could join?
A. IF YES:  If you could not get the rules changed, do you think you would resign from the club, even if your friends didn't?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/727/vshow}{GSS Data Explorer page for \code{racquit}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab 998 \tab 221 \tab 269 \tab 37 \tab 5 \tab 1530 \cr
   Total \tab 998 \tab 221 \tab 269 \tab 37 \tab 5 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racsups}},
\code{\link{racteach}}
}
\concept{Core}
\keyword{variable}
