% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racschol}
\alias{racschol}
\title{Should whites & blacks go to same school}
\source{
General Social Survey https://gss.norc.org
}
\description{
racschol
}
\details{
Question Was the high school you attended all black, integrated - but mostly black, or integrated - but mostly white?
}
\section{Values}{

\itemize{
\item \verb{[1]} same schools
\item \verb{[2]} separate schools
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1976, 1977, 1980, 1982, 1984, 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/394/vshow}{GSS Data Explorer page for \code{racschol}}.

Counts by year:\tabular{lllllll}{
   year \tab same schools \tab separate schools \tab don't know \tab no answer \tab iap \tab Total \cr
   1972 \tab 1381 \tab 193 \tab 32 \tab 7 \tab - \tab 1613 \cr
   1976 \tab 1250 \tab 207 \tab 39 \tab 3 \tab - \tab 1499 \cr
   1977 \tab 1298 \tab 197 \tab 25 \tab 10 \tab - \tab 1530 \cr
   1980 \tab 1273 \tab 160 \tab 29 \tab 6 \tab - \tab 1468 \cr
   1982 \tab 1670 \tab 141 \tab 43 \tab 6 \tab - \tab 1860 \cr
   1984 \tab 1327 \tab 110 \tab 26 \tab 10 \tab - \tab 1473 \cr
   1985 \tab 686 \tab 52 \tab 8 \tab 5 \tab 783 \tab 1534 \cr
   Total \tab 8885 \tab 1060 \tab 202 \tab 47 \tab 783 \tab 10977 \cr
}
}

\seealso{
Other Core: 
\code{\link{racclos}},
\code{\link{racdis}},
\code{\link{racfew}},
\code{\link{rachome}},
\code{\link{racinteg}},
\code{\link{raclive}},
\code{\link{racobjct}},
\code{\link{racopen}},
\code{\link{racseg}}
}
\concept{Core}
\keyword{variable}
