% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{racsubgv}
\alias{racsubgv}
\title{Suburban govts encourage integration}
\source{
General Social Survey https://gss.norc.org
}
\description{
racsubgv
}
\details{
Question What about the city government in white suburbs? Do you think they should encourage (Negro/Black) people to buy homes in the suburbs, should they discourage them, or should they leave it to private efforts?
}
\section{Values}{

\itemize{
\item \verb{[1]} encourage
\item \verb{[2]} discourage
\item \verb{[3]} leave alone
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/720/vshow}{GSS Data Explorer page for \code{racsubgv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab encourage \tab discourage \tab leave alone \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 193 \tab 146 \tab 1130 \tab 55 \tab 6 \tab 1530 \cr
   Total \tab 0 \tab 193 \tab 146 \tab 1130 \tab 55 \tab 6 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubs}}
}
\concept{Core}
\keyword{variable}
