% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{racsubs}
\alias{racsubs}
\title{Favor attempts to integrate suburbs}
\source{
General Social Survey https://gss.norc.org
}
\description{
racsubs
}
\details{
Question Some religious and business groups have set up programs to encourage (Negro/Black) people to buy houses in white suburbs. Do you favor or oppose these voluntary programs to integrate white suburbs?
}
\section{Values}{

\itemize{
\item \verb{[1]} favor
\item \verb{[2]} oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/719/vshow}{GSS Data Explorer page for \code{racsubs}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab favor \tab oppose \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 585 \tab 832 \tab 106 \tab 7 \tab 1530 \cr
   Total \tab 0 \tab 585 \tab 832 \tab 106 \tab 7 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubgv}}
}
\concept{Core}
\keyword{variable}
