% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{ractrust}
\alias{ractrust}
\title{Trust white people}
\source{
General Social Survey https://gss.norc.org
}
\description{
ractrust
}
\details{
Question Do you feel you can trust most white people, some white people, or no white people?
}
\section{Values}{

\itemize{
\item \code{1} most white people
\item \code{2} some white people
\item \code{3} no white people
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/409/vshow}{GSS Data Explorer page for \code{ractrust}}.
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}}
}
\concept{Core}
\keyword{variable}
