% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_133.R
\name{racwork}
\alias{racwork}
\title{Racial makeup of workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
racwork
}
\details{
Question Are the people who work where you work all white, mostly white, about half and half, mostly black, or all black?
}
\section{Values}{

\itemize{
\item \verb{[1]} all white
\item \verb{[2]} mostly white
\item \verb{[3]} about half and half
\item \verb{[4]} mostly black
\item \verb{[5]} all black
\item \verb{[6]} works alone
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   A/C/- \tab 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1249/vshow}{GSS Data Explorer page for \code{racwork}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab all white \tab mostly white \tab about half and half \tab mostly black \tab all black \tab works alone \tab don't know \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   1990 \tab 495 \tab 276 \tab 387 \tab 98 \tab 25 \tab 8 \tab 1 \tab 58 \tab 24 \tab - \tab - \tab 1372 \cr
   1996 \tab 2199 \tab 182 \tab 333 \tab 130 \tab 25 \tab 6 \tab - \tab 8 \tab 21 \tab - \tab - \tab 2904 \cr
   1998 \tab 1564 \tab 366 \tab 581 \tab 224 \tab 46 \tab 13 \tab - \tab 17 \tab 21 \tab - \tab - \tab 2832 \cr
   2000 \tab 1581 \tab 314 \tab 574 \tab 242 \tab 59 \tab 8 \tab - \tab 15 \tab 24 \tab - \tab - \tab 2817 \cr
   2002 \tab 2156 \tab 155 \tab 291 \tab 117 \tab 28 \tab 7 \tab - \tab 10 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 2220 \tab 163 \tab 259 \tab 131 \tab 25 \tab 3 \tab - \tab 6 \tab 5 \tab - \tab - \tab 2812 \cr
   2006 \tab 3264 \tab 290 \tab 548 \tab 307 \tab 52 \tab 7 \tab - \tab 33 \tab 9 \tab - \tab - \tab 4510 \cr
   2008 \tab 1198 \tab 168 \tab 336 \tab 229 \tab 49 \tab 8 \tab - \tab 30 \tab 5 \tab - \tab - \tab 2023 \cr
   2010 \tab 1284 \tab 159 \tab 358 \tab 167 \tab 37 \tab 4 \tab - \tab 30 \tab 5 \tab - \tab - \tab 2044 \cr
   2012 \tab 1196 \tab 167 \tab 333 \tab 192 \tab 42 \tab 5 \tab - \tab 22 \tab 17 \tab - \tab - \tab 1974 \cr
   2014 \tab 1504 \tab 194 \tab 443 \tab 321 \tab 42 \tab 5 \tab - \tab 23 \tab 6 \tab - \tab - \tab 2538 \cr
   2016 \tab 1735 \tab 210 \tab 483 \tab 340 \tab 42 \tab 11 \tab - \tab 27 \tab 19 \tab - \tab - \tab 2867 \cr
   2018 \tab 1389 \tab 160 \tab 387 \tab 291 \tab 59 \tab 6 \tab - \tab 41 \tab 15 \tab - \tab - \tab 2348 \cr
   2021 \tab 2539 \tab 163 \tab 757 \tab 397 \tab 49 \tab 5 \tab - \tab 78 \tab 30 \tab 14 \tab - \tab 4032 \cr
   2022 \tab 2196 \tab 146 \tab 580 \tab 456 \tab 67 \tab 11 \tab - \tab 71 \tab 14 \tab 3 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 26520 \tab 3113 \tab 6650 \tab 3642 \tab 647 \tab 107 \tab 1 \tab 469 \tab 216 \tab 17 \tab 3309 \tab 44691 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Race: 
\code{\link{admirblk}},
\code{\link{allowmar}},
\code{\link{asngovt}},
\code{\link{blkgains}},
\code{\link{blkgovt}},
\code{\link{blkinflu}},
\code{\link{blksimp}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{bwfriend}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{color}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{dsndntpay}},
\code{\link{ethchng}},
\code{\link{ethhist}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdmar}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{hseth}},
\code{\link{hspgovt}},
\code{\link{hswht}},
\code{\link{leastcom}},
\code{\link{meltpot}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{numbwfrd}},
\code{\link{oprace}},
\code{\link{owneth}},
\code{\link{perla}},
\code{\link{perla_self}},
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs16}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs7}},
\code{\link{raceacs9}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmeet}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpromo}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{ractrust}},
\code{\link{ratetone}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{trtmyrac}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wrkwayup}}
}
\concept{Intergroup Relations}
\concept{Race}
\keyword{variable}
