% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{radioact}
\alias{radioact}
\title{Sci knowledge: allradioactivity is man-made}
\source{
General Social Survey https://gss.norc.org
}
\description{
radioact
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don’t know or aren’t sure, just tell me so,  and we will skip to the next question. Remember true, false, or don’t know.
B. All radioactivity is man-made. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3335/vshow}{GSS Data Explorer page for \code{radioact}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 269 \tab 1292 \tab 302 \tab 1 \tab 4510 \cr
   2008 \tab 518 \tab 293 \tab 1026 \tab 176 \tab 10 \tab 2023 \cr
   2010 \tab 1103 \tab 175 \tab 625 \tab 139 \tab 2 \tab 2044 \cr
   2012 \tab 972 \tab 176 \tab 706 \tab 114 \tab 6 \tab 1974 \cr
   2014 \tab 1299 \tab 208 \tab 894 \tab 134 \tab 3 \tab 2538 \cr
   2016 \tab 1477 \tab 255 \tab 964 \tab 160 \tab 11 \tab 2867 \cr
   2018 \tab 1173 \tab 235 \tab 802 \tab 135 \tab 3 \tab 2348 \cr
   Total \tab 9188 \tab 1611 \tab 6309 \tab 1160 \tab 36 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{hotcore}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{viruses}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
