% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_328.R
\name{raised}
\alias{raised}
\title{Situation caused by the way he or she was raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
raised
}
\details{
Question E. In your opinion, how likely is it that (John/Mary)'s situation MIGHT be caused by the way he or she was raised?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2925/vshow}{GSS Data Explorer page for \code{raised}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 289 \tab 615 \tab 246 \tab 201 \tab 32 \tab 10 \tab 2765 \cr
   Total \tab 1372 \tab 289 \tab 615 \tab 246 \tab 201 \tab 32 \tab 10 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{allergic}},
\code{\link{bettrown}},
\code{\link{discipln}},
\code{\link{geneprob}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{nextdoor}},
\code{\link{violtv}}
}
\concept{Mental Health II}
\keyword{variable}
