% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_144.R
\name{raisehrd}
\alias{raisehrd}
\title{Workers who work hard and perform well got raise}
\source{
General Social Survey https://gss.norc.org
}
\description{
raisehrd
}
\details{
Question Think about how pay raises are given where you work. Please tell me all statements that apply. Do you feel that raises are given to...
B. Those workers who work hard and perform well.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1351/vshow}{GSS Data Explorer page for \code{raisehrd}}.
}

\seealso{
Other Work Organizations: 
\code{\link{fringe1}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{othpay}},
\code{\link{raiseall}},
\code{\link{raisefav}}

Other Job: 
\code{\link{fringe1}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{raiseall}},
\code{\link{raisefav}}
}
\concept{Job}
\concept{Work Organizations}
\keyword{variable}
