% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{random}
\alias{random}
\title{Randomization variable for form x}
\source{
General Social Survey https://gss.norc.org
}
\description{
random
}
\details{
Question Randomization for form
}
\section{Values}{

\itemize{
\item No applicable value labels.
\item 
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7092/vshow}{GSS Data Explorer page for \code{random}}.
}

\keyword{variable}
