% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{rank1}
\alias{rank1}
\title{R's self ranking of social position (2021)}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank1
}
\details{
Question In our society there are groups which tend to be towards the top and groups which tend to be towards the bottom. On the handcard is a scale that runs from top to bottom. Where would you put yourself now on this scale?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - top
\item \verb{[10]} 10 - bottom
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7512/vshow}{GSS Data Explorer page for \code{rank1}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 1 - top \tab 10 - bottom \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 34 \tab 40 \tab 44 \tab 228 \tab 267 \tab 730 \tab 197 \tab 165 \tab 78 \tab 29 \tab 5 \tab 161 \tab 25 \tab 4032 \cr
   Total \tab 2029 \tab 34 \tab 40 \tab 44 \tab 228 \tab 267 \tab 730 \tab 197 \tab 165 \tab 78 \tab 29 \tab 5 \tab 161 \tab 25 \tab 4032 \cr
}
}

\keyword{variable}
