% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{rank10}
\alias{rank10}
\title{R's social rank 10 years ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank10
}
\details{
Question In our society there are groups which tend to be towards the top and those that are towards the bottom. Here we have a scale that runs from top to bottom.
B. And ten years ago, where did you fit in then?
}
\section{Values}{

\itemize{
\item \code{1} 1 - top
\item \code{10} 10 - bottom
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/571/vshow}{GSS Data Explorer page for \code{rank10}}.
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()}

Other Social Class: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{rank}()}
}
\concept{Core}
\concept{Social Class}
\keyword{variable}
