% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{rank10fut}
\alias{rank10fut}
\title{R's social rank 10 years later}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank10fut
}
\details{
Question And thinking ahead 10 years from now, where do you think you will be on a scale of 1 to 10, where 1 is the top and 10 the bottom?
}
\section{Values}{

\itemize{
\item \code{1} 1 - top
\item \code{10} 10 - bottom
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7514/vshow}{GSS Data Explorer page for \code{rank10fut}}.
}

\keyword{variable}
