% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{rank16}
\alias{rank16}
\title{R's self ranking of family at age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank16
}
\details{
Question In our society there are groups which tend to be towards the top and groups which tend to be towards the bottom. (CATI TEXT: On the handcard is ) (WEB TEXT: Below is ) a scale that runs from top to bottom.) If you think about the family that you grew up in, where did they fit in then?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - top
\item \verb{[10]} 10 - bottom
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7513/vshow}{GSS Data Explorer page for \code{rank16}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 1 - top \tab 10 - bottom \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 34 \tab 55 \tab 44 \tab 160 \tab 226 \tab 599 \tab 243 \tab 237 \tab 146 \tab 64 \tab 4 \tab 163 \tab 28 \tab 4032 \cr
   Total \tab 2029 \tab 34 \tab 55 \tab 44 \tab 160 \tab 226 \tab 599 \tab 243 \tab 237 \tab 146 \tab 64 \tab 4 \tab 163 \tab 28 \tab 4032 \cr
}
}

\keyword{variable}
