% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_149.R
\name{rap}
\alias{rap}
\title{Like or dislike rap music}
\source{
General Social Survey https://gss.norc.org
}
\description{
rap
}
\details{
Question I'm going to read you a list of some types of music. Can you tell me which of the statements on this card comes closest to your feeling about each type of music (HAND CARD CA TO RESPONDENT.) Let's start with big band music. Do you like it very much, like it, have mixed feelings, dislike it, dislike it very much, or is this a type of music that you don't know much about?
N.  Rap music
}
\section{Values}{

\itemize{
\item \code{1} like very much
\item \code{2} like it
\item \code{3} mixed feelings
\item \code{4} dislike it
\item \code{5} dislike very much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1400/vshow}{GSS Data Explorer page for \code{rap}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab like very much \tab like it \tab mixed feelings \tab dislike it \tab dislike very much \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 44 \tab 159 \tab 284 \tab 433 \tab 614 \tab 61 \tab 11 \tab 1606 \cr
   Total \tab 0 \tab 44 \tab 159 \tab 284 \tab 433 \tab 614 \tab 61 \tab 11 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{conrock}},
\code{\link{hvymetal}},
\code{\link{makeart}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{reggae}},
\code{\link{visitart}}
}
\concept{Culture}
\keyword{variable}
