% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{ratechall124}
\alias{ratechall124}
\title{Rating of donald trump}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratechall124
}
\details{
Question How would you rate Donald Trump?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8091/vshow}{GSS Data Explorer page for \code{ratechall124}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 15 \tab 16 \tab 19 \tab 2 \tab 20 \tab 22 \tab 25 \tab 3 \tab 30 \tab 34 \tab 35 \tab 37 \tab 4 \tab 40 \tab 45 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 55 \tab 57 \tab 59 \tab 60 \tab 62 \tab 65 \tab 7 \tab 70 \tab 75 \tab 8 \tab 80 \tab 82 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 9 \tab 90 \tab 92 \tab 95 \tab 97 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1091 \tab 24 \tab 64 \tab 302 \tab 158 \tab 1 \tab 1 \tab 10 \tab 27 \tab 2 \tab 18 \tab 1 \tab 113 \tab 1 \tab 14 \tab 1 \tab 2 \tab 109 \tab 7 \tab 1 \tab 6 \tab 27 \tab 375 \tab 8 \tab 9 \tab 1 \tab 1 \tab 186 \tab 1 \tab 16 \tab 1 \tab 199 \tab 62 \tab 2 \tab 55 \tab 2 \tab 175 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 71 \tab 2 \tab 22 \tab 2 \tab 2 \tab 7 \tab 35 \tab 14 \tab 25 \tab 46 \tab 3309 \cr
   Total \tab 0 \tab 1091 \tab 24 \tab 64 \tab 302 \tab 158 \tab 1 \tab 1 \tab 10 \tab 27 \tab 2 \tab 18 \tab 1 \tab 113 \tab 1 \tab 14 \tab 1 \tab 2 \tab 109 \tab 7 \tab 1 \tab 6 \tab 27 \tab 375 \tab 8 \tab 9 \tab 1 \tab 1 \tab 186 \tab 1 \tab 16 \tab 1 \tab 199 \tab 62 \tab 2 \tab 55 \tab 2 \tab 175 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 71 \tab 2 \tab 22 \tab 2 \tab 2 \tab 7 \tab 35 \tab 14 \tab 25 \tab 46 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{ratechall224}},
\code{\link{ratechall324}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{ratemaga}},
\code{\link{raterepp}}
}
\concept{ANES}
\keyword{variable}
