% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{ratechall3ts}
\alias{ratechall3ts}
\title{Rating of presidential challenger 3 - time series}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratechall3ts
}
\details{
Question (Recode of RATECHALL324)
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8096/vshow}{GSS Data Explorer page for \code{ratechall3ts}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 22 \tab 23 \tab 25 \tab 3 \tab 30 \tab 32 \tab 33 \tab 35 \tab 37 \tab 39 \tab 40 \tab 43 \tab 44 \tab 45 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 55 \tab 6 \tab 60 \tab 65 \tab 67 \tab 69 \tab 70 \tab 71 \tab 72 \tab 75 \tab 80 \tab 82 \tab 85 \tab 90 \tab 95 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 459 \tab 9 \tab 51 \tab 46 \tab 3 \tab 154 \tab 1 \tab 1 \tab 5 \tab 41 \tab 3 \tab 1 \tab 25 \tab 4 \tab 170 \tab 1 \tab 1 \tab 10 \tab 1 \tab 1 \tab 160 \tab 1 \tab 1 \tab 19 \tab 3 \tab 5 \tab 25 \tab 994 \tab 1 \tab 1 \tab 15 \tab 2 \tab 218 \tab 15 \tab 1 \tab 1 \tab 136 \tab 1 \tab 2 \tab 32 \tab 38 \tab 1 \tab 71 \tab 13 \tab 1 \tab 1 \tab 2 \tab 444 \tab 11 \tab 25 \tab 82 \tab 3309 \cr
   Total \tab 0 \tab 459 \tab 9 \tab 51 \tab 46 \tab 3 \tab 154 \tab 1 \tab 1 \tab 5 \tab 41 \tab 3 \tab 1 \tab 25 \tab 4 \tab 170 \tab 1 \tab 1 \tab 10 \tab 1 \tab 1 \tab 160 \tab 1 \tab 1 \tab 19 \tab 3 \tab 5 \tab 25 \tab 994 \tab 1 \tab 1 \tab 15 \tab 2 \tab 218 \tab 15 \tab 1 \tab 1 \tab 136 \tab 1 \tab 2 \tab 32 \tab 38 \tab 1 \tab 71 \tab 13 \tab 1 \tab 1 \tab 2 \tab 444 \tab 11 \tab 25 \tab 82 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{ratechall1ts}},
\code{\link{ratechall2ts}},
\code{\link{rateincumbts}},
\code{\link{voteprob}}
}
\concept{ANES}
\keyword{variable}
