% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{ratedemp}
\alias{ratedemp}
\title{Rating of democratic party}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratedemp
}
\details{
Question How would you rate the Democratic Party?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8097/vshow}{GSS Data Explorer page for \code{ratedemp}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 15 \tab 16 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 23 \tab 24 \tab 25 \tab 27 \tab 3 \tab 30 \tab 33 \tab 35 \tab 4 \tab 40 \tab 45 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 55 \tab 56 \tab 57 \tab 6 \tab 60 \tab 65 \tab 66 \tab 69 \tab 7 \tab 70 \tab 71 \tab 72 \tab 75 \tab 76 \tab 77 \tab 78 \tab 8 \tab 80 \tab 85 \tab 86 \tab 88 \tab 9 \tab 90 \tab 95 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 508 \tab 18 \tab 77 \tab 132 \tab 3 \tab 173 \tab 3 \tab 1 \tab 1 \tab 7 \tab 69 \tab 1 \tab 1 \tab 1 \tab 39 \tab 1 \tab 3 \tab 214 \tab 3 \tab 18 \tab 2 \tab 210 \tab 16 \tab 2 \tab 1 \tab 4 \tab 23 \tab 602 \tab 1 \tab 4 \tab 34 \tab 1 \tab 1 \tab 2 \tab 285 \tab 39 \tab 3 \tab 1 \tab 3 \tab 248 \tab 1 \tab 1 \tab 65 \tab 1 \tab 1 \tab 1 \tab 2 \tab 64 \tab 173 \tab 1 \tab 1 \tab 1 \tab 41 \tab 17 \tab 1 \tab 2 \tab 81 \tab 14 \tab 26 \tab 60 \tab 3309 \cr
   Total \tab 0 \tab 508 \tab 18 \tab 77 \tab 132 \tab 3 \tab 173 \tab 3 \tab 1 \tab 1 \tab 7 \tab 69 \tab 1 \tab 1 \tab 1 \tab 39 \tab 1 \tab 3 \tab 214 \tab 3 \tab 18 \tab 2 \tab 210 \tab 16 \tab 2 \tab 1 \tab 4 \tab 23 \tab 602 \tab 1 \tab 4 \tab 34 \tab 1 \tab 1 \tab 2 \tab 285 \tab 39 \tab 3 \tab 1 \tab 3 \tab 248 \tab 1 \tab 1 \tab 65 \tab 1 \tab 1 \tab 1 \tab 2 \tab 64 \tab 173 \tab 1 \tab 1 \tab 1 \tab 41 \tab 17 \tab 1 \tab 2 \tab 81 \tab 14 \tab 26 \tab 60 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{aneseligible}},
\code{\link{anespost_status}},
\code{\link{anespre_status}},
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovote24a}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
