% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{rateincumbts}
\alias{rateincumbts}
\title{Rating of incumbent president - time series}
\source{
General Social Survey https://gss.norc.org
}
\description{
rateincumbts
}
\details{
Question (Recode of RATEINCUMB24)
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8099/vshow}{GSS Data Explorer page for \code{rateincumbts}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 13 \tab 15 \tab 18 \tab 2 \tab 20 \tab 22 \tab 23 \tab 25 \tab 29 \tab 3 \tab 30 \tab 32 \tab 34 \tab 35 \tab 4 \tab 40 \tab 45 \tab 46 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 55 \tab 57 \tab 58 \tab 6 \tab 60 \tab 61 \tab 65 \tab 68 \tab 70 \tab 73 \tab 75 \tab 80 \tab 85 \tab 88 \tab 89 \tab 90 \tab 91 \tab 95 \tab 98 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 764 \tab 25 \tab 69 \tab 176 \tab 2 \tab 1 \tab 205 \tab 1 \tab 10 \tab 66 \tab 1 \tab 1 \tab 33 \tab 2 \tab 4 \tab 206 \tab 1 \tab 2 \tab 8 \tab 3 \tab 185 \tab 19 \tab 1 \tab 4 \tab 6 \tab 26 \tab 457 \tab 2 \tab 13 \tab 1 \tab 1 \tab 1 \tab 235 \tab 1 \tab 16 \tab 1 \tab 234 \tab 2 \tab 68 \tab 63 \tab 194 \tab 2 \tab 2 \tab 49 \tab 1 \tab 20 \tab 2 \tab 6 \tab 34 \tab 14 \tab 22 \tab 47 \tab 3309 \cr
   Total \tab 0 \tab 764 \tab 25 \tab 69 \tab 176 \tab 2 \tab 1 \tab 205 \tab 1 \tab 10 \tab 66 \tab 1 \tab 1 \tab 33 \tab 2 \tab 4 \tab 206 \tab 1 \tab 2 \tab 8 \tab 3 \tab 185 \tab 19 \tab 1 \tab 4 \tab 6 \tab 26 \tab 457 \tab 2 \tab 13 \tab 1 \tab 1 \tab 1 \tab 235 \tab 1 \tab 16 \tab 1 \tab 234 \tab 2 \tab 68 \tab 63 \tab 194 \tab 2 \tab 2 \tab 49 \tab 1 \tab 20 \tab 2 \tab 6 \tab 34 \tab 14 \tab 22 \tab 47 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{aneseligible}},
\code{\link{anespost_status}},
\code{\link{anespre_status}},
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovote24a}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
