% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{ratemaga}
\alias{ratemaga}
\title{Rating of maga republicans}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratemaga
}
\details{
Question How would you rate “MAGA” Republicans?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8100/vshow}{GSS Data Explorer page for \code{ratemaga}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 15 \tab 2 \tab 20 \tab 25 \tab 30 \tab 32 \tab 35 \tab 4 \tab 40 \tab 41 \tab 42 \tab 45 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 55 \tab 6 \tab 60 \tab 65 \tab 7 \tab 70 \tab 75 \tab 78 \tab 79 \tab 8 \tab 80 \tab 82 \tab 85 \tab 86 \tab 88 \tab 90 \tab 95 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1084 \tab 20 \tab 64 \tab 182 \tab 1 \tab 167 \tab 7 \tab 42 \tab 20 \tab 142 \tab 1 \tab 11 \tab 1 \tab 116 \tab 1 \tab 1 \tab 10 \tab 2 \tab 26 \tab 545 \tab 2 \tab 1 \tab 8 \tab 1 \tab 136 \tab 19 \tab 1 \tab 119 \tab 39 \tab 1 \tab 1 \tab 1 \tab 47 \tab 1 \tab 118 \tab 2 \tab 2 \tab 47 \tab 13 \tab 4 \tab 189 \tab 16 \tab 28 \tab 70 \tab 3309 \cr
   Total \tab 0 \tab 1084 \tab 20 \tab 64 \tab 182 \tab 1 \tab 167 \tab 7 \tab 42 \tab 20 \tab 142 \tab 1 \tab 11 \tab 1 \tab 116 \tab 1 \tab 1 \tab 10 \tab 2 \tab 26 \tab 545 \tab 2 \tab 1 \tab 8 \tab 1 \tab 136 \tab 19 \tab 1 \tab 119 \tab 39 \tab 1 \tab 1 \tab 1 \tab 47 \tab 1 \tab 118 \tab 2 \tab 2 \tab 47 \tab 13 \tab 4 \tab 189 \tab 16 \tab 28 \tab 70 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{ratechall124}},
\code{\link{ratechall224}},
\code{\link{ratechall324}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{raterepp}}
}
\concept{ANES}
\keyword{variable}
