% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{ratepain}
\alias{ratepain}
\title{In the past 7 days, r's average pain}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratepain
}
\details{
Question On a scale from 0 to 10, with 0 meaning no pain and 10 being the worst imaginable pain, how would you rate your pain on average?
}
\section{Values}{

\itemize{
\item \verb{[0]} 0 - no pain
\item \verb{[10]} 10 - the worst imaginable pain
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7171/vshow}{GSS Data Explorer page for \code{ratepain}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 0 - no pain \tab 1 \tab 10 - the worst imaginable pain \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab - \tab 569 \tab 353 \tab 23 \tab 312 \tab 295 \tab 193 \tab 204 \tab 143 \tab 140 \tab 68 \tab 28 \tab 3 \tab 17 \tab 2348 \cr
   Total \tab 0 \tab 569 \tab 353 \tab 23 \tab 312 \tab 295 \tab 193 \tab 204 \tab 143 \tab 140 \tab 68 \tab 28 \tab 3 \tab 17 \tab 2348 \cr
}
}

\keyword{variable}
