% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{ratepain1}
\alias{ratepain1}
\title{In the past 7 days, r's average pain (2021)}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratepain1
}
\details{
Question On a scale from 0 to 10, with 0 meaning no pain and 10 being the worst imaginable pain, how would you rate your pain on average?
}
\section{Values}{

\itemize{
\item \verb{[0]} 0 - no pain
\item \verb{[10]} 10 - the worst imaginable pain
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7521/vshow}{GSS Data Explorer page for \code{ratepain1}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 0 - no pain \tab 1 \tab 10 - the worst imaginable pain \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 389 \tab 828 \tab 620 \tab 21 \tab 600 \tab 430 \tab 258 \tab 302 \tab 211 \tab 173 \tab 114 \tab 56 \tab 2 \tab 1 \tab 27 \tab 4032 \cr
   Total \tab 389 \tab 828 \tab 620 \tab 21 \tab 600 \tab 430 \tab 258 \tab 302 \tab 211 \tab 173 \tab 114 \tab 56 \tab 2 \tab 1 \tab 27 \tab 4032 \cr
}
}

\keyword{variable}
