% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{raterepp}
\alias{raterepp}
\title{Rating of republican party}
\source{
General Social Survey https://gss.norc.org
}
\description{
raterepp
}
\details{
Question How would you rate the Republican Party?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8101/vshow}{GSS Data Explorer page for \code{raterepp}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 15 \tab 16 \tab 2 \tab 20 \tab 22 \tab 25 \tab 3 \tab 30 \tab 32 \tab 35 \tab 38 \tab 4 \tab 40 \tab 45 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 54 \tab 55 \tab 56 \tab 6 \tab 60 \tab 63 \tab 65 \tab 66 \tab 7 \tab 70 \tab 72 \tab 75 \tab 78 \tab 79 \tab 8 \tab 80 \tab 82 \tab 84 \tab 85 \tab 87 \tab 88 \tab 89 \tab 90 \tab 95 \tab 99 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 573 \tab 12 \tab 82 \tab 131 \tab 2 \tab 212 \tab 1 \tab 12 \tab 78 \tab 1 \tab 52 \tab 4 \tab 219 \tab 1 \tab 24 \tab 1 \tab 2 \tab 215 \tab 19 \tab 2 \tab 3 \tab 31 \tab 591 \tab 1 \tab 2 \tab 2 \tab 22 \tab 2 \tab 3 \tab 239 \tab 1 \tab 22 \tab 2 \tab 1 \tab 231 \tab 1 \tab 65 \tab 1 \tab 1 \tab 4 \tab 61 \tab 1 \tab 1 \tab 134 \tab 1 \tab 1 \tab 1 \tab 38 \tab 13 \tab 1 \tab 91 \tab 14 \tab 26 \tab 58 \tab 3309 \cr
   Total \tab 0 \tab 573 \tab 12 \tab 82 \tab 131 \tab 2 \tab 212 \tab 1 \tab 12 \tab 78 \tab 1 \tab 52 \tab 4 \tab 219 \tab 1 \tab 24 \tab 1 \tab 2 \tab 215 \tab 19 \tab 2 \tab 3 \tab 31 \tab 591 \tab 1 \tab 2 \tab 2 \tab 22 \tab 2 \tab 3 \tab 239 \tab 1 \tab 22 \tab 2 \tab 1 \tab 231 \tab 1 \tab 65 \tab 1 \tab 1 \tab 4 \tab 61 \tab 1 \tab 1 \tab 134 \tab 1 \tab 1 \tab 1 \tab 38 \tab 13 \tab 1 \tab 91 \tab 14 \tab 26 \tab 58 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{ratechall124}},
\code{\link{ratechall224}},
\code{\link{ratechall324}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{ratemaga}}
}
\concept{ANES}
\keyword{variable}
