% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{ratetone}
\alias{ratetone}
\title{R's facial coloring by interviewer}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratetone
}
\details{
Question INTERVIEWER: PLEASE RECORD THE COLOR FROM THE COLOR CARD THAT MOST CLOSELY CORRESPONDS TO THE RESPONDENT'S FACIAL COLORING.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - lightest
\item \verb{[10]} 10 - darkest
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3978/vshow}{GSS Data Explorer page for \code{ratetone}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - lightest \tab 10 - darkest \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab not available in this year \tab Total \cr
   2012 \tab - \tab 561 \tab 18 \tab 531 \tab 234 \tab 106 \tab 72 \tab 45 \tab 65 \tab 49 \tab 17 \tab 276 \tab - \tab 1974 \cr
   2014 \tab - \tab 1045 \tab 6 \tab 513 \tab 247 \tab 102 \tab 86 \tab 60 \tab 58 \tab 39 \tab 28 \tab 354 \tab - \tab 2538 \cr
   2016 \tab - \tab 1049 \tab 9 \tab 785 \tab 292 \tab 166 \tab 116 \tab 82 \tab 79 \tab 61 \tab 21 \tab 207 \tab - \tab 2867 \cr
   2018 \tab - \tab 904 \tab 3 \tab 556 \tab 219 \tab 106 \tab 86 \tab 60 \tab 56 \tab 46 \tab 17 \tab 295 \tab - \tab 2348 \cr
   2022 \tab 1864 \tab 785 \tab 9 \tab 353 \tab 175 \tab 93 \tab 75 \tab 65 \tab 60 \tab 52 \tab 13 \tab - \tab - \tab 3544 \cr
   Total \tab 1864 \tab 4344 \tab 45 \tab 2738 \tab 1167 \tab 573 \tab 435 \tab 312 \tab 318 \tab 247 \tab 96 \tab 1132 \tab 0 \tab 13271 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidneedy}},
\code{\link{aidold}},
\code{\link{aidsmart}},
\code{\link{aidunemp}},
\code{\link{beltup}},
\code{\link{bribe}},
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{buspow}},
\code{\link{bustax}},
\code{\link{cctv}},
\code{\link{colop}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{cutdebt}},
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{databank}},
\code{\link{demworks}},
\code{\link{emonitor}},
\code{\link{eqincome}},
\code{\link{equalize}},
\code{\link{exploit}},
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{fewtrsty}},
\code{\link{gendereq}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{govtinfo}},
\code{\link{govtpow}},
\code{\link{grnlaws}},
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}},
\code{\link{hlphitec}},
\code{\link{hlpinflu}},
\code{\link{hlthcare}},
\code{\link{hsbasics}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hsrespct}},
\code{\link{hssci}},
\code{\link{hssexed}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{jobsall}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidout}},
\code{\link{kidskips}},
\code{\link{kidxfilm}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{mustret}},
\code{\link{nosmoke}},
\code{\link{obeylaw}},
\code{\link{opfamily}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{polgbeco}},
\code{\link{polint}},
\code{\link{polsfair}},
\code{\link{prespop}},
\code{\link{pricecon}},
\code{\link{progtax}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{protstrs}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{prvdschl}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{racists}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revoltrs}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}},
\code{\link{unionsok}},
\code{\link{verdict}},
\code{\link{volacty2}},
\code{\link{volactyr}},
\code{\link{wotrial}},
\code{\link{youinflu}}

Other Race: 
\code{\link{admirblk}},
\code{\link{allowmar}},
\code{\link{asngovt}},
\code{\link{blkgains}},
\code{\link{blkgovt}},
\code{\link{blkinflu}},
\code{\link{blksimp}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{bwfriend}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{color}},
\code{\link{congeth}},
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{dsndntpay}},
\code{\link{ethchng}},
\code{\link{ethhist}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{forbdmar}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{hispanic}},
\code{\link{hispanic_0022}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{hseth}},
\code{\link{hspgovt}},
\code{\link{hswht}},
\code{\link{leastcom}},
\code{\link{meltpot}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{numbwfrd}},
\code{\link{oprace}},
\code{\link{owneth}},
\code{\link{perla}},
\code{\link{perla_self}},
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs16}},
\code{\link{raceacs3}},
\code{\link{raceacs5}},
\code{\link{raceacs7}},
\code{\link{raceacs9}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmeet}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpromo}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{ractrust}},
\code{\link{racwork}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{trtmyrac}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wrkwayup}}
}
\concept{ISSP Role of Gov't}
\concept{Race}
\keyword{variable}
