% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_94.R
\name{rclose2}
\alias{rclose2}
\title{R feels especially close to person #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
rclose2
}
\details{
Question Do you feel equally close to all these people?
A. Which of these people do you feel especially close to?  REFER TO MATRIX ON PAGE 338 FOR NAME NUMBERS. CODE ALL RESPONDENT FEELS ESPECIALLY CLOSE TO, IF ONLY ONE MENTIONED,
PROBE:  Anyone else?
2. Second name
}
\section{Values}{

\itemize{
\item \code{1} especially close
\item \code{2} not esp. close
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/851/vshow}{GSS Data Explorer page for \code{rclose2}}.
}

\seealso{
Other Social Networks: 
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{rclose1}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}}

Other Social Networks: 
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{rclose1}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}}
}
\concept{Social Networks}
\keyword{variable}
