% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_94.R
\name{rclose3}
\alias{rclose3}
\title{R feels especially close to person #3}
\source{
General Social Survey https://gss.norc.org
}
\description{
rclose3
}
\details{
Question Do you feel equally close to all these people?
A. Which of these people do you feel especially close to?  REFER TO MATRIX ON PAGE 338 FOR NAME NUMBERS. CODE ALL RESPONDENT FEELS ESPECIALLY CLOSE TO, IF ONLY ONE MENTIONED,
PROBE:  Anyone else?
3. Third name
}
\section{Values}{

\itemize{
\item \code{1} especially close
\item \code{2} not esp. close
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/852/vshow}{GSS Data Explorer page for \code{rclose3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab especially close \tab not esp. close \tab no answer \tab Total \cr
   1985 \tab 1070 \tab 130 \tab 329 \tab 5 \tab 1534 \cr
   1987 \tab 1345 \tab 62 \tab 372 \tab 40 \tab 1819 \cr
   Total \tab 2415 \tab 192 \tab 701 \tab 45 \tab 3353 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose4}},
\code{\link{rclose5}}
}
\concept{Social Networks}
\keyword{variable}
