% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_94.R
\name{rclose5}
\alias{rclose5}
\title{R feels especially close to person #5}
\source{
General Social Survey https://gss.norc.org
}
\description{
rclose5
}
\details{
Question Do you feel equally close to all these people?
A. Which of these people do you feel especially close to?  REFER TO MATRIX ON PAGE 338 FOR NAME NUMBERS. CODE ALL RESPONDENT FEELS ESPECIALLY CLOSE TO, IF ONLY ONE MENTIONED,
PROBE:  Anyone else?
5. Fifth name
}
\section{Values}{

\itemize{
\item \verb{[1]} especially close
\item \verb{[2]} not esp. close
\item \verb{[7]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/854/vshow}{GSS Data Explorer page for \code{rclose5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab especially close \tab not esp. close \tab no answer \tab Total \cr
   1985 \tab 1328 \tab 54 \tab 148 \tab 4 \tab 1534 \cr
   Total \tab 1328 \tab 54 \tab 148 \tab 4 \tab 1534 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}}
}
\concept{Social Networks}
\keyword{variable}
