% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{rdeemlib}
\alias{rdeemlib}
\title{Contrast between redeemer and liberator}
\source{
General Social Survey https://gss.norc.org
}
\description{
rdeemlib
}
\details{
Question There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? (HAND CARD AE) The first set of contrasting images shows Mother at 1 on the scale and Father at 7. If you imagine God as a Mother you would place yourself at 1. If you imagine God as a Father, you would place yourself at 7. If you imagine God as somewhere between Mother and Father, you would place yourself at 2, 3, 4, 5, or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F, SUBSTITUTING IMAGES A-F FOR "MOTHER" AND "FATHER".) Where would you place your image of God on the scale for . . . READ EACH SET OF IMAGES AND CIRCLE ONE CODE FOR EACH.
}
\section{Values}{

\itemize{
\item \code{1} redeemer
\item \code{7} liberator
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
   1985 \tab - \tab full \cr
   1986 \tab - \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/339/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab liberator \tab no answer \tab redeemer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 78 \tab 47 \tab 342 \tab 25 \tab 18 \tab 2 \tab 46 \tab 56 \tab 375 \tab - \tab 1473 \cr
   1985 \tab - \tab 109 \tab 101 \tab 553 \tab 41 \tab 33 \tab 21 \tab 94 \tab 58 \tab 524 \tab - \tab 1534 \cr
   1986 \tab - \tab 101 \tab 72 \tab 535 \tab 47 \tab 25 \tab 16 \tab 78 \tab 50 \tab 546 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 64853 \tab 288 \tab 220 \tab 1430 \tab 113 \tab 76 \tab 39 \tab 218 \tab 164 \tab 1445 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}}

Other Religion: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}}

Other Split Ballots: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}}
}
\concept{Core}
\concept{Religion}
\concept{Split Ballots}
\keyword{variable}
