% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{rdiscaff}
\alias{rdiscaff}
\title{Chances family member not promoted for other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
rdiscaff
}
\details{
Question A. What do you think the chances are these days that you or anyone in your family won't get a job or promotion while an equally or less qualified (White/(Black/African-American)) employee receives one instead?  Would you say the chances of this happening are very likely, somewhat likely, somewhat unlikely, or very unlikely?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1647/vshow}{GSS Data Explorer page for \code{rdiscaff}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 213 \tab 458 \tab 340 \tab 358 \tab 88 \tab 17 \tab 2992 \cr
   Total \tab 1518 \tab 213 \tab 458 \tab 340 \tab 358 \tab 88 \tab 17 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rimmdisc}}
}
\concept{Multiculturalism}
\keyword{variable}
