% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{reactok}
\alias{reactok}
\title{Did r feel reaction was appropriate}
\source{
General Social Survey https://gss.norc.org
}
\description{
reactok
}
\details{
Question Did you feel like your reaction to the situation was appropriate, or did it seem wrong to you somehow... for example, too intense or the wrong emotion for the occasion? On a scale that runs from 0 as completely right to 10 as completely wrong for the situation, tell me whether your feelings were right or wrong for the situation.
}
\section{Values}{

\itemize{
\item \code{0} completely right
\item \code{10} completely wrong
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1783/vshow}{GSS Data Explorer page for \code{reactok}}.
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}

Other Feelings: 
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{othresp}},
\code{\link{selfresp}}
}
\concept{Emotions}
\concept{Feelings}
\keyword{variable}
