% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{readfict}
\alias{readfict}
\title{Read novels poems or plays}
\source{
General Social Survey https://gss.norc.org
}
\description{
readfict
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
A. Read novels, short stories, poems, or plays, other than those required by work or school.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2129/vshow}{GSS Data Explorer page for \code{readfict}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1998 \tab 1387 \tab 968 \tab 466 \tab 11 \tab - \tab 2832 \cr
   2002 \tab 1393 \tab 987 \tab 371 \tab 6 \tab 8 \tab 2765 \cr
   Total \tab 2780 \tab 1955 \tab 837 \tab 17 \tab 8 \tab 5597 \cr
}
}

\seealso{
Other Culture: 
\code{\link{drama}},
\code{\link{popmusic}}

Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{drama}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{popmusic}},
\code{\link{relart}},
\code{\link{stateart}},
\code{\link{volarts}}
}
\concept{Culture}
\concept{Religion and Culture}
\keyword{variable}
