% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{readword}
\alias{readword}
\title{How often did r read bible last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
readword
}
\details{
Question Have you read any part of the Bible at home within the last year.
Yes..........(ASK A)...............1
No….…................................2
IF YES: A. How often have you read the Bible in the last year? (USE CATEGORIES AS PROBES IF NEEDED).
}
\section{Values}{

\itemize{
\item \verb{[1]} several x day
\item \verb{[2]} once a day
\item \verb{[3]} several x week
\item \verb{[4]} once a week
\item \verb{[5]} lt once a week
\item \verb{[6]} didnt read last year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1080/vshow}{GSS Data Explorer page for \code{readword}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab several x day \tab once a day \tab several x week \tab once a week \tab lt once a week \tab didnt read last year \tab no answer \tab Total \cr
   1988 \tab 4 \tab 49 \tab 135 \tab 157 \tab 130 \tab 446 \tab 557 \tab 3 \tab 1481 \cr
   1998 \tab 1387 \tab 37 \tab 116 \tab 143 \tab 132 \tab 406 \tab 602 \tab 9 \tab 2832 \cr
   Total \tab 1391 \tab 86 \tab 251 \tab 300 \tab 262 \tab 852 \tab 1159 \tab 12 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{blkwhite}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{gochurch}},
\code{\link{permoral}},
\code{\link{punsin}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
