% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{realrinc}
\alias{realrinc}
\title{R's income in constant $}
\source{
General Social Survey https://gss.norc.org
}
\description{
realrinc
}
\details{
Question Respondent's income on 1972-2006 surveys in constant dollars (base = 1986)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Income: 
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{incdef}},
\code{\link{minfour}},
\code{\link{povline}},
\code{\link{realinc}}
}
\concept{Income}
\keyword{variable}
