% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{realtora}
\alias{realtora}
\title{R's own previous knowledge made r choose}
\source{
General Social Survey https://gss.norc.org
}
\description{
realtora
}
\details{
Question What made you choose the realtor with whom you worked? (Choose all that apply.)
A. My own previous knowledge of the realtor/agency.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1818/vshow}{GSS Data Explorer page for \code{realtora}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2686 \tab 51 \tab 114 \tab 2 \tab 51 \tab 2904 \cr
   Total \tab 2686 \tab 51 \tab 114 \tab 2 \tab 51 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
