% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{realtorb}
\alias{realtorb}
\title{A friend-relative recommended r choose}
\source{
General Social Survey https://gss.norc.org
}
\description{
realtorb
}
\details{
Question What made you choose the realtor with whom you worked? (Choose all that apply.)
B. A friend or relative recommended them.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1819/vshow}{GSS Data Explorer page for \code{realtorb}}.
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}

Other Purchases/Transactions: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\concept{Purchases/Transactions}
\keyword{variable}
