% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_194.R
\name{realtore}
\alias{realtore}
\title{I saw their advertisement made r choose}
\source{
General Social Survey https://gss.norc.org
}
\description{
realtore
}
\details{
Question What made you choose the realtor with whom you worked? (Choose all that apply.)
E. I saw their advertisement.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1822/vshow}{GSS Data Explorer page for \code{realtore}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2686 \tab 72 \tab 106 \tab 2 \tab 38 \tab 2904 \cr
   Total \tab 2686 \tab 72 \tab 106 \tab 2 \tab 38 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{lawyer}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{realtor1}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{sathome}},
\code{\link{warrntyh}}
}
\concept{Markets}
\keyword{variable}
