% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_194.R
\name{realtorf}
\alias{realtorf}
\title{Convenience made r choose the realtor}
\source{
General Social Survey https://gss.norc.org
}
\description{
realtorf
}
\details{
Question What made you choose the realtor with whom you worked? (Choose all that apply.)
F. Convenience.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1823/vshow}{GSS Data Explorer page for \code{realtorf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2686 \tab 53 \tab 112 \tab 2 \tab 51 \tab 2904 \cr
   Total \tab 2686 \tab 53 \tab 112 \tab 2 \tab 51 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{lawyer}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{realtor1}},
\code{\link{realtore}},
\code{\link{relagent}},
\code{\link{sathome}},
\code{\link{warrntyh}}
}
\concept{Markets}
\keyword{variable}
