% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{recycle}
\alias{recycle}
\title{Recycle cans bottles}
\source{
General Social Survey https://gss.norc.org
}
\description{
recycle
}
\details{
Question A. How often do you make a special effort to sort glass or cans or plastic or papers and so on for recycling?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{5} recycling not available where i live
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4769/vshow}{GSS Data Explorer page for \code{recycle}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab never \tab recycling not available where i live \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 579 \tab 338 \tab 343 \tab 204 \tab 55 \tab 1 \tab 37 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 481 \tab 329 \tab 340 \tab 170 \tab 47 \tab - \tab 19 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 393 \tab 291 \tab 271 \tab 189 \tab 67 \tab 2 \tab 63 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 563 \tab 324 \tab 317 \tab 190 \tab 28 \tab 1 \tab 7 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 1006 \tab 278 \tab 218 \tab 198 \tab 130 \tab - \tab 192 \tab 7 \tab 4032 \cr
   Total \tab 5813 \tab 3022 \tab 1560 \tab 1489 \tab 951 \tab 327 \tab 4 \tab 318 \tab 7 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
