% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{redcehme}
\alias{redcehme}
\title{How often reduce fuel}
\source{
General Social Survey https://gss.norc.org
}
\description{
redcehme
}
\details{
Question K1. How often do you reduce the energy or fuel you use at home for environmental reasons?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4813/vshow}{GSS Data Explorer page for \code{redcehme}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   2010 \tab 614 \tab 165 \tab 440 \tab 450 \tab 362 \tab 6 \tab 7 \tab 2044 \cr
   Total \tab 614 \tab 165 \tab 440 \tab 450 \tab 362 \tab 6 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{peopgrn}},
\code{\link{tempgen1}}
}
\concept{ISSP Environment}
\keyword{variable}
