% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_499.R
\name{reducemp}
\alias{reducemp}
\title{Importance of educ. & formal qual.}
\source{
General Social Survey https://gss.norc.org
}
\description{
reducemp
}
\details{
Question Think of two people doing the same kind of work in your place. What Do you personally think is important in deciding
Looking at the things below, please write ''1'' in the box next to the thing that you think is most important at your place of work. Then write ''2'' next to the thing that you think is next most important, and ''3'' next to the thing that you think is third most important. Leave the other boxes blank.
At my work, deciding on two people doing the same kind of work, how important is...
G. The employee''s education and formal qualifications
}
\section{Values}{

\itemize{
\item \verb{[1]} most impt
\item \verb{[2]} 2nd most impt
\item \verb{[3]} 3rd most impt
\item \verb{[4]} not chosen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4539/vshow}{GSS Data Explorer page for \code{reducemp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab most impt \tab 2nd most impt \tab 3rd most impt \tab not chosen \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1989 \tab 639 \tab 51 \tab 82 \tab 151 \tab 417 \tab 25 \tab 172 \tab - \tab 1537 \cr
   Total \tab 639 \tab 51 \tab 82 \tab 151 \tab 417 \tab 25 \tab 172 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{adforjob}},
\code{\link{ageemp}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{bosswrks}},
\code{\link{chngewrk}},
\code{\link{chngwork}},
\code{\link{cowrkers}},
\code{\link{dailywrk}},
\code{\link{danger}},
\code{\link{decidwrk}},
\code{\link{discwk5}},
\code{\link{dk}},
\code{\link{dowell}},
\code{\link{econsup}},
\code{\link{econsup1}},
\code{\link{edcskill}},
\code{\link{educemp}},
\code{\link{employer}},
\code{\link{empself}},
\code{\link{evjob}},
\code{\link{exjobsat}},
\code{\link{expernc}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{famintjb}},
\code{\link{famorjob}},
\code{\link{famresp}},
\code{\link{findjob}},
\code{\link{findwork}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{getjob}},
\code{\link{harass5}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrsmoney}},
\code{\link{hwduties}},
\code{\link{indusoth}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{joborfam}},
\code{\link{jobsat}},
\code{\link{jobskill}},
\code{\link{leavejob}},
\code{\link{leisure}},
\code{\link{localnum}},
\code{\link{lookjob}},
\code{\link{lowpay}},
\code{\link{moonlite}},
\code{\link{mortravl}},
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{newjob}},
\code{\link{newskill}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{otherwrk}},
\code{\link{othhrs}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paysame}},
\code{\link{paytime}},
\code{\link{physwrk}},
\code{\link{placewrk}},
\code{\link{planwrk}},
\code{\link{priagncy}},
\code{\link{prideorg}},
\code{\link{privgovt}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{pubagncy}},
\code{\link{rageemp}},
\code{\link{rcontact}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{replaceu}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rimpskls}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpaysame}},
\code{\link{rpromotn}},
\code{\link{rsecjob}},
\code{\link{rwrkindp}},
\code{\link{ryrsfirm}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexemp}},
\code{\link{skltrain}},
\code{\link{smallbig}},
\code{\link{spwrkgvt}},
\code{\link{stayorg3}},
\code{\link{stress}},
\code{\link{strngun}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{tempwork}},
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timeoff}},
\code{\link{timepdwk}},
\code{\link{timerelx}},
\code{\link{totalnum}},
\code{\link{unbetter}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unfamten}},
\code{\link{unhlthy}},
\code{\link{unionsbd}},
\code{\link{unjobexp}},
\code{\link{unjobsec}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unpleznt}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{useskill}},
\code{\link{wantads}},
\code{\link{wantjob}},
\code{\link{wantjob1}},
\code{\link{wantjob2}},
\code{\link{whydisc5}},
\code{\link{whyjbend}},
\code{\link{wkathome}},
\code{\link{wkpersnl}},
\code{\link{wktenure}},
\code{\link{wkwkends}},
\code{\link{worecsup}},
\code{\link{worknow}},
\code{\link{workweek}},
\code{\link{wornojob}},
\code{\link{worryjob}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}},
\code{\link{wrkindp}},
\code{\link{wrkshift}},
\code{\link{xhaustn}},
\code{\link{yrjobend}},
\code{\link{yrsfirm}}

Other Employment: 
\code{\link{accntsci}},
\code{\link{advances}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compwage}},
\code{\link{curempyr}},
\code{\link{denyrais}},
\code{\link{dowell}},
\code{\link{drugtest}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empyears}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{evunemp}},
\code{\link{expernc}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findjob}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{futpromo}},
\code{\link{harjob5}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{ignorwk}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobyears}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{knowschd}},
\code{\link{laidoff}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leavejb}},
\code{\link{liedcwkr}},
\code{\link{losejb12}},
\code{\link{losejob5}},
\code{\link{lostjobs}},
\code{\link{ltrsref}},
\code{\link{marrcoun}},
\code{\link{medtreat}},
\code{\link{moredays}},
\code{\link{moretrde}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{neverwk}},
\code{\link{newjobs}},
\code{\link{numemp}},
\code{\link{numemply}},
\code{\link{numemps}},
\code{\link{numemps5}},
\code{\link{numemps_0422}},
\code{\link{numown}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{ownbiz}},
\code{\link{paidemps}},
\code{\link{partfull}},
\code{\link{paysame}},
\code{\link{physical}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{ryrsfirm}},
\code{\link{samejob}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{skiltest}},
\code{\link{slsmnshp}},
\code{\link{spwrkgvt}},
\code{\link{startbiz}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totunemp}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{whywkhme}},
\code{\link{wkbhvrs}},
\code{\link{work3yrs}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrkstat}},
\code{\link{wrktype}},
\code{\link{wrkyears}},
\code{\link{yearbiz}},
\code{\link{yearsjob}},
\code{\link{yrsfirm}}

Other Job: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{actupset}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{baseofop}},
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{bosswrks}},
\code{\link{chngeorg}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compperf}},
\code{\link{condemnd}},
\code{\link{cowrkers}},
\code{\link{cowrkhlp}},
\code{\link{cowrkhrd}},
\code{\link{cowrkint}},
\code{\link{cowrkrel}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{dailywrk}},
\code{\link{danger}},
\code{\link{datefrst}},
\code{\link{denyrais}},
\code{\link{deptperf}},
\code{\link{difstand}},
\code{\link{discaff}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{discaffy}},
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{donothng}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{exptrain}},
\code{\link{extr2001}},
\code{\link{extr2005}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{famintjb}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{fehire}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findnwjb}},
\code{\link{findout}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{formltrn}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{fringeok}},
\code{\link{ftwork}},
\code{\link{ftwork1}},
\code{\link{gdjobsec}},
\code{\link{genderwk}},
\code{\link{handmove}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{haveinfo}},
\code{\link{helporg}},
\code{\link{hiinc}},
\code{\link{hlpequip}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hlthtype}},
\code{\link{hotargus}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{hrsmoney}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{imatter}},
\code{\link{inclstjb}},
\code{\link{indperf}},
\code{\link{indusoth}},
\code{\link{intjob}},
\code{\link{intlcowk}},
\code{\link{iwrkhard}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobaff}},
\code{\link{jobcntct}},
\code{\link{jobcntrl}},
\code{\link{jobcowrk}},
\code{\link{jobeasy}},
\code{\link{jobfrnds}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobinfo}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobpay}},
\code{\link{jobplace}},
\code{\link{jobpromo}},
\code{\link{jobpub}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobsecok}},
\code{\link{jobskill}},
\code{\link{jobtvrad}},
\code{\link{jobvshme}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{knowschd}},
\code{\link{knowwhat}},
\code{\link{lackinfo}},
\code{\link{lastjob}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{learning}},
\code{\link{learnjob}},
\code{\link{learnnew}},
\code{\link{leavejb}},
\code{\link{leavejob}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{localnum}},
\code{\link{locyrs}},
\code{\link{lookaway}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{lotofsay}},
\code{\link{lottosay}},
\code{\link{manvsemp}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{misswork}},
\code{\link{monitred}},
\code{\link{monthly}},
\code{\link{moonlite}},
\code{\link{moretrde}},
\code{\link{mykids}},
\code{\link{myskills}},
\code{\link{negjob5}},
\code{\link{newjob}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{ntwkhard}},
\code{\link{numemp}},
\code{\link{numemps5}},
\code{\link{numorg}},
\code{\link{numown}},
\code{\link{numsites}},
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{offsup}},
\code{\link{opdevel}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{orgsize}},
\code{\link{othcredt}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{otherwrk}},
\code{\link{othhrs}},
\code{\link{othplan}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{overwork}},
\code{\link{owninc}},
\code{\link{ownstock}},
\code{\link{paidemps}},
\code{\link{parleave}},
\code{\link{partorg}},
\code{\link{partteam}},
\code{\link{pastpay}},
\code{\link{pawrkslf}},
\code{\link{paychnge}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{perspace}},
\code{\link{phyeffrt}},
\code{\link{physharm}},
\code{\link{physwrk}},
\code{\link{placewrk}},
\code{\link{planwrk}},
\code{\link{privgovt}},
\code{\link{prodctiv}},
\code{\link{promotn}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{proudorg}},
\code{\link{putdown}},
\code{\link{rageemp}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{rdiscaff}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{reliedon}},
\code{\link{replaceu}},
\code{\link{respect}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rimmdisc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpaysame}},
\code{\link{rpromotn}},
\code{\link{rptprobs}},
\code{\link{rsecjob}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{rwrkindp}},
\code{\link{ryrsfirm}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{samevals}},
\code{\link{samework}},
\code{\link{satjob}},
\code{\link{satjob1}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{schoolng}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{secondwk}},
\code{\link{seecowrk}},
\code{\link{sethours}},
\code{\link{setthngs}},
\code{\link{shout}},
\code{\link{siteyrs}},
\code{\link{skipwork}},
\code{\link{slsmnshp}},
\code{\link{smallbig}},
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{splocnum}},
\code{\link{spvtrfair}},
\code{\link{spwrkslf}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{stress}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supervis}},
\code{\link{suphelp}},
\code{\link{supnum}},
\code{\link{supothrs}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{teamsafe}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timeoff}},
\code{\link{toldwork}},
\code{\link{toofewwk}},
\code{\link{totalnum}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{training}},
\code{\link{trainops}},
\code{\link{trainsch}},
\code{\link{trdestck}},
\code{\link{treatres}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{typeorg}},
\code{\link{unhlthy}},
\code{\link{unmanrel}},
\code{\link{unpleznt}},
\code{\link{useskill}},
\code{\link{usetech}},
\code{\link{wantjob}},
\code{\link{wantjob1}},
\code{\link{weekly}},
\code{\link{wherewrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkbhvrs}},
\code{\link{wkbonus}},
\code{\link{wkcomptr}},
\code{\link{wkcontct}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkpraise}},
\code{\link{wksmooth}},
\code{\link{wksub}},
\code{\link{wksup}},
\code{\link{wktenure}},
\code{\link{wktopsat}},
\code{\link{wocomptr}},
\code{\link{work3yrs}},
\code{\link{workdiff}},
\code{\link{workdy}},
\code{\link{workfast}},
\code{\link{workless}},
\code{\link{worryjob}},
\code{\link{wrkindp}},
\code{\link{wrkoth}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrktime}},
\code{\link{xhaustn}},
\code{\link{yearbiz}},
\code{\link{yearly}},
\code{\link{yearsjob}},
\code{\link{yearval}},
\code{\link{youmoney}},
\code{\link{yousup}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}
}
\concept{Employment}
\concept{Job}
\concept{Occupation}
\concept{Work Orientation}
\keyword{variable}
