% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{refmorwk}
\alias{refmorwk}
\title{Take on additional paid work?}
\source{
General Social Survey https://gss.norc.org
}
\description{
refmorwk
}
\details{
Question In your (present/most recent) job, (have you ever done/did you ever do) any of the following because of your responsibilities to members of your family?
B. Take on additional paid work?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1668/vshow}{GSS Data Explorer page for \code{refmorwk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab don't know \tab skipped on web \tab Total \cr
   1996 \tab 1444 \tab 518 \tab 824 \tab 118 \tab - \tab - \tab - \tab 2904 \cr
   2024 \tab 1162 \tab 832 \tab 1268 \tab 6 \tab - \tab 15 \tab 26 \tab 3309 \cr
   Total \tab 2606 \tab 1350 \tab 2092 \tab 124 \tab 0 \tab 15 \tab 26 \tab 6213 \cr
}
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{earnmore}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}

Other Family: 
\code{\link{balwkfam}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}}

Other Job: 
\code{\link{noathome}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}}

Other Work: 
\code{\link{balwkfam}},
\code{\link{noathome}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}
}
\concept{Family}
\concept{Gender}
\concept{Job}
\concept{Work}
\keyword{variable}
