% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{refrndms}
\alias{refrndms}
\title{Referendum are good way to decide important pol quest}
\source{
General Social Survey https://gss.norc.org
}
\description{
refrndms
}
\details{
Question Thinking now about politics in America, to what extent do you agree or disagree with the following statements?
C. Referendums are a good way to decide important political questions.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4954/vshow}{GSS Data Explorer page for \code{refrndms}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 161 \tab 715 \tab 379 \tab 106 \tab 17 \tab 86 \tab 8 \tab 2812 \cr
   2014 \tab 1274 \tab 73 \tab 491 \tab 369 \tab 101 \tab 25 \tab 197 \tab 8 \tab 2538 \cr
   Total \tab 2614 \tab 234 \tab 1206 \tab 748 \tab 207 \tab 42 \tab 283 \tab 16 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
