% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{refxhour}
\alias{refxhour}
\title{Refuse to work overtime or extra hours?}
\source{
General Social Survey https://gss.norc.org
}
\description{
refxhour
}
\details{
Question In your (present/most recent) job, (have you ever done/did you ever do) any of the following because of your responsibilities to members of your family?
C. Refuse to work overtime or extra hours?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1669/vshow}{GSS Data Explorer page for \code{refxhour}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1444 \tab 302 \tab 1041 \tab 117 \tab - \tab 2904 \cr
   Total \tab 1444 \tab 302 \tab 1041 \tab 117 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{balwkfam}},
\code{\link{children}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{earnmore}},
\code{\link{emoteoth}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{febear}},
\code{\link{fehire}},
\code{\link{feimp}},
\code{\link{fejobaff}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{flextime}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{haprom}},
\code{\link{hmemaker}},
\code{\link{livewith}},
\code{\link{manprof}},
\code{\link{marfree}},
\code{\link{mebear}},
\code{\link{men}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{parleave}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{rhmewrk}},
\code{\link{romance}},
\code{\link{rushed}},
\code{\link{selffrst}},
\code{\link{sexhar}},
\code{\link{sexpromo}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{twoincs}},
\code{\link{willwed1}},
\code{\link{willwed2}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{wrkclass}},
\code{\link{yourself}}

Other Family: 
\code{\link{adoptkid}},
\code{\link{agewed}},
\code{\link{agewedcr}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{balwkfam}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{chldben}},
\code{\link{chlddth}},
\code{\link{chldhrt}},
\code{\link{clsclsex}},
\code{\link{cousins}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce}},
\code{\link{engda}},
\code{\link{engson}},
\code{\link{evdiv}},
\code{\link{evkid}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{famlife}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famvswk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{feovrwrk}},
\code{\link{flextime}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{inlawnum}},
\code{\link{inperson}},
\code{\link{jobvsfam}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidcall}},
\code{\link{kidlived}},
\code{\link{kidmost}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{kidvisit}},
\code{\link{letters}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{madeath}},
\code{\link{malive}},
\code{\link{malive1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{marital}},
\code{\link{marunion}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{meetings}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{mykids}},
\code{\link{niecenep}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{numcntct}},
\code{\link{numprobs}},
\code{\link{obtohelp}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{othkids}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{padeath}},
\code{\link{palive}},
\code{\link{palive1}},
\code{\link{parleave}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{pawrkgrw}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{posslq}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relcall}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relmarry}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{rfamlook}},
\code{\link{sibcall}},
\code{\link{sibdeath}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibvisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{spdeath}},
\code{\link{spfalook}},
\code{\link{spfam16}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{sppart}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{widowed}},
\code{\link{wkvsfam}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}}

Other Job: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{actupset}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{baseofop}},
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{bosswrks}},
\code{\link{chngeorg}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compperf}},
\code{\link{condemnd}},
\code{\link{cowrkers}},
\code{\link{cowrkhlp}},
\code{\link{cowrkhrd}},
\code{\link{cowrkint}},
\code{\link{cowrkrel}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{dailywrk}},
\code{\link{danger}},
\code{\link{datefrst}},
\code{\link{denyrais}},
\code{\link{deptperf}},
\code{\link{difstand}},
\code{\link{discaff}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{discaffy}},
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{donothng}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{exptrain}},
\code{\link{extr2001}},
\code{\link{extr2005}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{famintjb}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{fehire}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findnwjb}},
\code{\link{findout}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{formltrn}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{fringeok}},
\code{\link{ftwork}},
\code{\link{ftwork1}},
\code{\link{gdjobsec}},
\code{\link{genderwk}},
\code{\link{handmove}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{haveinfo}},
\code{\link{helporg}},
\code{\link{hiinc}},
\code{\link{hlpequip}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hlthtype}},
\code{\link{hotargus}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{hrsmoney}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{imatter}},
\code{\link{inclstjb}},
\code{\link{indperf}},
\code{\link{indusoth}},
\code{\link{intjob}},
\code{\link{intlcowk}},
\code{\link{iwrkhard}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobaff}},
\code{\link{jobcntct}},
\code{\link{jobcntrl}},
\code{\link{jobcowrk}},
\code{\link{jobeasy}},
\code{\link{jobfrnds}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobinfo}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobpay}},
\code{\link{jobplace}},
\code{\link{jobpromo}},
\code{\link{jobpub}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobsecok}},
\code{\link{jobskill}},
\code{\link{jobtvrad}},
\code{\link{jobvshme}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{knowschd}},
\code{\link{knowwhat}},
\code{\link{lackinfo}},
\code{\link{lastjob}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{learning}},
\code{\link{learnjob}},
\code{\link{learnnew}},
\code{\link{leavejb}},
\code{\link{leavejob}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{localnum}},
\code{\link{locyrs}},
\code{\link{lookaway}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{lotofsay}},
\code{\link{lottosay}},
\code{\link{manvsemp}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{misswork}},
\code{\link{monitred}},
\code{\link{monthly}},
\code{\link{moonlite}},
\code{\link{moretrde}},
\code{\link{mykids}},
\code{\link{myskills}},
\code{\link{negjob5}},
\code{\link{newjob}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{ntwkhard}},
\code{\link{numemp}},
\code{\link{numemps5}},
\code{\link{numorg}},
\code{\link{numown}},
\code{\link{numsites}},
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{offsup}},
\code{\link{opdevel}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{orgsize}},
\code{\link{othcredt}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{otherwrk}},
\code{\link{othhrs}},
\code{\link{othplan}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{overwork}},
\code{\link{owninc}},
\code{\link{ownstock}},
\code{\link{paidemps}},
\code{\link{parleave}},
\code{\link{partorg}},
\code{\link{partteam}},
\code{\link{pastpay}},
\code{\link{pawrkslf}},
\code{\link{paychnge}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{perspace}},
\code{\link{phyeffrt}},
\code{\link{physharm}},
\code{\link{physwrk}},
\code{\link{placewrk}},
\code{\link{planwrk}},
\code{\link{privgovt}},
\code{\link{prodctiv}},
\code{\link{promotn}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{proudorg}},
\code{\link{putdown}},
\code{\link{rageemp}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{rdiscaff}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{reliedon}},
\code{\link{replaceu}},
\code{\link{respect}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rimmdisc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpaysame}},
\code{\link{rpromotn}},
\code{\link{rptprobs}},
\code{\link{rsecjob}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{rwrkindp}},
\code{\link{ryrsfirm}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{samevals}},
\code{\link{samework}},
\code{\link{satjob}},
\code{\link{satjob1}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{schoolng}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{secondwk}},
\code{\link{seecowrk}},
\code{\link{sethours}},
\code{\link{setthngs}},
\code{\link{shout}},
\code{\link{siteyrs}},
\code{\link{skipwork}},
\code{\link{slsmnshp}},
\code{\link{smallbig}},
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{splocnum}},
\code{\link{spvtrfair}},
\code{\link{spwrkslf}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{stress}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supervis}},
\code{\link{suphelp}},
\code{\link{supnum}},
\code{\link{supothrs}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{teamsafe}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timeoff}},
\code{\link{toldwork}},
\code{\link{toofewwk}},
\code{\link{totalnum}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{training}},
\code{\link{trainops}},
\code{\link{trainsch}},
\code{\link{trdestck}},
\code{\link{treatres}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{typeorg}},
\code{\link{unhlthy}},
\code{\link{unmanrel}},
\code{\link{unpleznt}},
\code{\link{useskill}},
\code{\link{usetech}},
\code{\link{wantjob}},
\code{\link{wantjob1}},
\code{\link{weekly}},
\code{\link{wherewrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkbhvrs}},
\code{\link{wkbonus}},
\code{\link{wkcomptr}},
\code{\link{wkcontct}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkpraise}},
\code{\link{wksmooth}},
\code{\link{wksub}},
\code{\link{wksup}},
\code{\link{wktenure}},
\code{\link{wktopsat}},
\code{\link{wocomptr}},
\code{\link{work3yrs}},
\code{\link{workdiff}},
\code{\link{workdy}},
\code{\link{workfast}},
\code{\link{workless}},
\code{\link{worryjob}},
\code{\link{wrkindp}},
\code{\link{wrkoth}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrktime}},
\code{\link{xhaustn}},
\code{\link{yearbiz}},
\code{\link{yearly}},
\code{\link{yearsjob}},
\code{\link{yearval}},
\code{\link{youmoney}},
\code{\link{yousup}}

Other Work: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{balwkfam}},
\code{\link{buyvalue}},
\code{\link{chngwork}},
\code{\link{cmprgmng}},
\code{\link{commute}},
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{denyrais}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{empinput}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{extr2005}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{getahead}},
\code{\link{getaheay}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{impjob}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jobcntrl}},
\code{\link{jobfind}},
\code{\link{jobfind1}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobpromo}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobskill}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{lottosay}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{monitred}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{numorg}},
\code{\link{offsup}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{parleave}},
\code{\link{pawrkslf}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{phyeffrt}},
\code{\link{placewrk}},
\code{\link{prideorg}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{richwork}},
\code{\link{rimpskls}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{satjob}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexhar}},
\code{\link{skltrain}},
\code{\link{slfmangd}},
\code{\link{slsmnshp}},
\code{\link{spdays}},
\code{\link{sphedslf}},
\code{\link{sphour}},
\code{\link{spvtrfair}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{spwrkslf}},
\code{\link{stayorg3}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{talkteam}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{toldwork}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{trynewjb}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{useskill}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkcontct}},
\code{\link{wkfambiz}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkracism}},
\code{\link{wkrspns}},
\code{\link{wksexism}},
\code{\link{wksub}},
\code{\link{wksubs}},
\code{\link{wksup}},
\code{\link{wksups}},
\code{\link{wktenure}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{workdy}},
\code{\link{workfor}},
\code{\link{workhr}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{worknow}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkhome}},
\code{\link{wrkimp}},
\code{\link{wrkindep}},
\code{\link{wrkindp}},
\code{\link{wrkmuch}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{wrkwell}},
\code{\link{yearsjob}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Family}
\concept{Gender}
\concept{Job}
\concept{Work}
\keyword{variable}
