% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{reg16}
\alias{reg16}
\title{Region of residence, age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
reg16
}
\details{
Question In what state or foreign country were you living when you were 16 years old?
As of 2024, REG16 has been recoded to only account for four categories. This has been retroactively applied to all previous years of REG16. The original REG16 containing data from 1972 to 2022 has been renamed REG16_7222.
}
\section{Values}{

\itemize{
\item \verb{[0]} foreign
\item \verb{[1]} northeast
\item \verb{[2]} midwest
\item \verb{[3]} south
\item \verb{[4]} west
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/84/vshow}{GSS Data Explorer page for \code{reg16}}.

Counts by year:\tabular{llllllllll}{
   year \tab foreign \tab northeast \tab midwest \tab south \tab west \tab iap \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 66 \tab 388 \tab 460 \tab 524 \tab 175 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 58 \tab 317 \tab 427 \tab 531 \tab 171 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 54 \tab 312 \tab 465 \tab 491 \tab 162 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 50 \tab 315 \tab 436 \tab 519 \tab 170 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 66 \tab 327 \tab 458 \tab 483 \tab 165 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 73 \tab 312 \tab 469 \tab 499 \tab 177 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 61 \tab 336 \tab 464 \tab 508 \tab 163 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 64 \tab 290 \tab 433 \tab 496 \tab 185 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 73 \tab 353 \tab 521 \tab 726 \tab 187 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 76 \tab 363 \tab 505 \tab 468 \tab 187 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 62 \tab 307 \tab 451 \tab 470 \tab 183 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 66 \tab 315 \tab 432 \tab 511 \tab 210 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 72 \tab 307 \tab 451 \tab 445 \tab 195 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 67 \tab 340 \tab 523 \tab 688 \tab 201 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 76 \tab 303 \tab 443 \tab 462 \tab 197 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 82 \tab 312 \tab 459 \tab 464 \tab 220 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 70 \tab 290 \tab 411 \tab 406 \tab 195 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 66 \tab 323 \tab 427 \tab 493 \tab 208 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 84 \tab 301 \tab 468 \tab 515 \tab 238 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 166 \tab 608 \tab 774 \tab 991 \tab 453 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 180 \tab 631 \tab 776 \tab 881 \tab 436 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 174 \tab 597 \tab 748 \tab 910 \tab 403 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 191 \tab 599 \tab 714 \tab 891 \tab 422 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 195 \tab 591 \tab 761 \tab 837 \tab 381 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 194 \tab 503 \tab 738 \tab 905 \tab 472 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 489 \tab 756 \tab 1086 \tab 1418 \tab 761 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 193 \tab 372 \tab 503 \tab 602 \tab 353 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 189 \tab 370 \tab 514 \tab 630 \tab 341 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 205 \tab 350 \tab 480 \tab 608 \tab 331 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 280 \tab 477 \tab 591 \tab 742 \tab 448 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 249 \tab 530 \tab 737 \tab 865 \tab 486 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 217 \tab 383 \tab 555 \tab 774 \tab 419 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 282 \tab 719 \tab 1147 \tab 1127 \tab 743 \tab 14 \tab - \tab - \tab 4032 \cr
   2022 \tab 315 \tab 570 \tab 855 \tab 1147 \tab 647 \tab - \tab 1 \tab 9 \tab 3544 \cr
   2024 \tab 311 \tab 577 \tab 868 \tab 1012 \tab 524 \tab - \tab 3 \tab 14 \tab 3309 \cr
   Total \tab 5116 \tab 14744 \tab 20550 \tab 24039 \tab 11209 \tab 14 \tab 4 \tab 23 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\concept{Region}
\keyword{variable}
