% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{regdoc}
\alias{regdoc}
\title{R has regular doctor or clinic}
\source{
General Social Survey https://gss.norc.org
}
\description{
regdoc
}
\details{
Question This section is about physicians telling patients about how health insurance and managed care policies affect health care. First, I will ask you about your health care. Later, you will hear a recorded conversation between a doctor and patient and I'll ask some questions about the conversation.Do you have a regular doctor or clinic?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2738/vshow}{GSS Data Explorer page for \code{regdoc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 2294 \tab 460 \tab 7 \tab 4 \tab 2765 \cr
   Total \tab 0 \tab 2294 \tab 460 \tab 7 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{docvig}},
\code{\link{regdocyr}}
}
\concept{Doctors & Patients}
\keyword{variable}
