% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{regdocyr}
\alias{regdocyr}
\title{R has had doctor for more than one year}
\source{
General Social Survey https://gss.norc.org
}
\description{
regdocyr
}
\details{
Question Have you been a patient with this doctor or clinic for more than one year?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2739/vshow}{GSS Data Explorer page for \code{regdocyr}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{docvig}},
\code{\link{regdoc}}
}
\concept{Doctors & Patients}
\keyword{variable}
