% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{region}
\alias{region}
\title{Region of interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
region
}
\details{
Question A. REGION OF INTERVIEW
}
\section{Values}{

\itemize{
\item \verb{[1]} northeast
\item \verb{[2]} midwest
\item \verb{[3]} south
\item \verb{[4]} west
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/119/vshow}{GSS Data Explorer page for \code{region}}.

Counts by year:\tabular{llllll}{
   year \tab northeast \tab midwest \tab south \tab west \tab Total \cr
   1972 \tab 388 \tab 446 \tab 499 \tab 280 \tab 1613 \cr
   1973 \tab 343 \tab 428 \tab 480 \tab 253 \tab 1504 \cr
   1974 \tab 328 \tab 429 \tab 477 \tab 250 \tab 1484 \cr
   1975 \tab 326 \tab 429 \tab 500 \tab 235 \tab 1490 \cr
   1976 \tab 340 \tab 424 \tab 481 \tab 254 \tab 1499 \cr
   1977 \tab 301 \tab 456 \tab 508 \tab 265 \tab 1530 \cr
   1978 \tab 335 \tab 443 \tab 502 \tab 252 \tab 1532 \cr
   1980 \tab 292 \tab 408 \tab 498 \tab 270 \tab 1468 \cr
   1982 \tab 394 \tab 512 \tab 676 \tab 278 \tab 1860 \cr
   1983 \tab 339 \tab 484 \tab 488 \tab 288 \tab 1599 \cr
   1984 \tab 295 \tab 416 \tab 495 \tab 267 \tab 1473 \cr
   1985 \tab 295 \tab 410 \tab 537 \tab 292 \tab 1534 \cr
   1986 \tab 303 \tab 404 \tab 487 \tab 276 \tab 1470 \cr
   1987 \tab 341 \tab 506 \tab 680 \tab 292 \tab 1819 \cr
   1988 \tab 289 \tab 407 \tab 504 \tab 281 \tab 1481 \cr
   1989 \tab 303 \tab 401 \tab 529 \tab 304 \tab 1537 \cr
   1990 \tab 273 \tab 377 \tab 458 \tab 264 \tab 1372 \cr
   1991 \tab 299 \tab 380 \tab 544 \tab 294 \tab 1517 \cr
   1993 \tab 290 \tab 417 \tab 559 \tab 340 \tab 1606 \cr
   1994 \tab 566 \tab 715 \tab 1095 \tab 616 \tab 2992 \cr
   1996 \tab 569 \tab 699 \tab 1016 \tab 620 \tab 2904 \cr
   1998 \tab 569 \tab 696 \tab 1021 \tab 546 \tab 2832 \cr
   2000 \tab 568 \tab 670 \tab 1006 \tab 573 \tab 2817 \cr
   2002 \tab 576 \tab 701 \tab 957 \tab 531 \tab 2765 \cr
   2004 \tab 449 \tab 697 \tab 1084 \tab 582 \tab 2812 \cr
   2006 \tab 711 \tab 1038 \tab 1745 \tab 1016 \tab 4510 \cr
   2008 \tab 355 \tab 469 \tab 750 \tab 449 \tab 2023 \cr
   2010 \tab 338 \tab 474 \tab 787 \tab 445 \tab 2044 \cr
   2012 \tab 335 \tab 457 \tab 742 \tab 440 \tab 1974 \cr
   2014 \tab 435 \tab 573 \tab 926 \tab 604 \tab 2538 \cr
   2016 \tab 488 \tab 695 \tab 1052 \tab 632 \tab 2867 \cr
   2018 \tab 356 \tab 513 \tab 953 \tab 526 \tab 2348 \cr
   2021 \tab 617 \tab 990 \tab 1496 \tab 929 \tab 4032 \cr
   2022 \tab 525 \tab 798 \tab 1415 \tab 806 \tab 3544 \cr
   2024 \tab 541 \tab 812 \tab 1310 \tab 646 \tab 3309 \cr
   Total \tab 14072 \tab 19174 \tab 27257 \tab 15196 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{dotdata}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom98}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}

Other Residence: 
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Core}
\concept{Region}
\concept{Residence}
\keyword{variable}
