% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{reincar}
\alias{reincar}
\title{Believe in reincarnation}
\source{
General Social Survey https://gss.norc.org
}
\description{
reincar
}
\details{
Question Do you believe in…
A. ...in Reincarnation – being reborn in this world again and again?
}
\section{Values}{

\itemize{
\item \code{1} yes, definitely
\item \code{2} yes, probably
\item \code{3} no, probably not
\item \code{4} no, definitely not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4705/vshow}{GSS Data Explorer page for \code{reincar}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably not \tab no, definitely not \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 178 \tab 248 \tab 374 \tab 488 \tab 71 \tab 6 \tab 2023 \cr
   Total \tab 658 \tab 178 \tab 248 \tab 374 \tab 488 \tab 71 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{ancestrs}},
\code{\link{mywaygod}},
\code{\link{nirvana}},
\code{\link{relcand}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\keyword{variable}
