% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{relactiv1}
\alias{relactiv1}
\title{How often does r take part in relig activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
relactiv1
}
\details{
Question How often do you take part in the activities and organizations of a church or place of worship other than attending services?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than once a year
\item \verb{[3]} about once or twice a year
\item \verb{[4]} several times a year
\item \verb{[5]} about once a month
\item \verb{[6]} 2-3 times a month
\item \verb{[7]} nearly every week
\item \verb{[8]} every week
\item \verb{[9]} several times a week
\item \verb{[10]} once a day
\item \verb{[11]} several times a day
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4678/vshow}{GSS Data Explorer page for \code{relactiv1}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab never \tab less than once a year \tab about once or twice a year \tab several times a year \tab about once a month \tab 2-3 times a month \tab nearly every week \tab every week \tab several times a week \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 549 \tab 112 \tab 202 \tab 165 \tab 72 \tab 87 \tab 27 \tab 74 \tab 68 \tab 4 \tab 5 \tab 2023 \cr
   Total \tab 658 \tab 549 \tab 112 \tab 202 \tab 165 \tab 72 \tab 87 \tab 27 \tab 74 \tab 68 \tab 4 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{feelrel}},
\code{\link{prayfreq}},
\code{\link{relactiv}},
\code{\link{schlpray}}
}
\concept{ISSP Religion}
\keyword{variable}
