% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{relalt1}
\alias{relalt1}
\title{Rel change altered r's outlook on life}
\source{
General Social Survey https://gss.norc.org
}
\description{
relalt1
}
\details{
Question To what extent did the religious/spiritual change you just told me about do the following?
A. Altered your outlook on life?
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{7} a great deal
\item \code{9} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3077/vshow}{GSS Data Explorer page for \code{relalt1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab not at all \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab a great deal \tab refused \tab Total \cr
   2004 \tab 2163 \tab 19 \tab 9 \tab 19 \tab 60 \tab 104 \tab 117 \tab 310 \tab 11 \tab 2812 \cr
   Total \tab 2163 \tab 19 \tab 9 \tab 19 \tab 60 \tab 104 \tab 117 \tab 310 \tab 11 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{alloflfe}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}}
}
\concept{Spirituality and Religion}
\keyword{variable}
