% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{relalt2}
\alias{relalt2}
\title{Rel change made r sense life was meaningful}
\source{
General Social Survey https://gss.norc.org
}
\description{
relalt2
}
\details{
Question To what extent did the religious/spiritual change you just told me about do the following?
B. Gave you a greater sense that your life was meaningful and significant
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[7]} a great deal
\item \verb{[9]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3078/vshow}{GSS Data Explorer page for \code{relalt2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab not at all \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab a great deal \tab refused \tab Total \cr
   2004 \tab 2163 \tab 19 \tab 9 \tab 24 \tab 59 \tab 74 \tab 126 \tab 327 \tab 11 \tab 2812 \cr
   Total \tab 2163 \tab 19 \tab 9 \tab 24 \tab 59 \tab 74 \tab 126 \tab 327 \tab 11 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{alloflfe}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{relalt1}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}}
}
\concept{Spirituality and Religion}
\keyword{variable}
