% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{relalt3}
\alias{relalt3}
\title{Rel change changed goals in life}
\source{
General Social Survey https://gss.norc.org
}
\description{
relalt3
}
\details{
Question To what extent did the religious/spiritual change you just told me about do the following?
C. Changed your goals in life
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[7]} a great deal
\item \verb{[9]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3079/vshow}{GSS Data Explorer page for \code{relalt3}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab a great deal \tab 8 \tab refused \tab Total \cr
   2004 \tab 2163 \tab 72 \tab 28 \tab 35 \tab 86 \tab 111 \tab 84 \tab 219 \tab 3 \tab 11 \tab 2812 \cr
   Total \tab 2163 \tab 72 \tab 28 \tab 35 \tab 86 \tab 111 \tab 84 \tab 219 \tab 3 \tab 11 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{alloflfe}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}}
}
\concept{Spirituality and Religion}
\keyword{variable}
