% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{relalt5}
\alias{relalt5}
\title{Rel change changed day to day activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
relalt5
}
\details{
Question To what extent did the religious/spiritual change you just told me about do the following?
E. Changed your day-to-day activities
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[7]} a great deal
\item \verb{[9]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3081/vshow}{GSS Data Explorer page for \code{relalt5}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab a great deal \tab 8 \tab refused \tab Total \cr
   2004 \tab 2163 \tab 84 \tab 34 \tab 55 \tab 102 \tab 113 \tab 83 \tab 165 \tab 1 \tab 12 \tab 2812 \cr
   Total \tab 2163 \tab 84 \tab 34 \tab 55 \tab 102 \tab 113 \tab 83 \tab 165 \tab 1 \tab 12 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{alloflfe}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt6}}
}
\concept{Spirituality and Religion}
\keyword{variable}
