% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{relart}
\alias{relart}
\title{Had experience of art at place of worship}
\source{
General Social Survey https://gss.norc.org
}
\description{
relart
}
\details{
Question Within the past twelve months, have you had any experiences with the arts at your place of worship?  I mean activities like viewing performances or exhibits, singing in a choir, or playing an instrument.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2132/vshow}{GSS Data Explorer page for \code{relart}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 1797 \tab 313 \tab 643 \tab 79 \tab - \tab 2832 \cr
   Total \tab 1797 \tab 313 \tab 643 \tab 79 \tab 0 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{beausprt}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{drama}},
\code{\link{effort}},
\code{\link{feelgod}},
\code{\link{fidgety}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godlove}},
\code{\link{godwatch}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{lesspain}},
\code{\link{localart}},
\code{\link{meditate}},
\code{\link{natarts}},
\code{\link{natlart}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{popmusic}},
\code{\link{privpray}},
\code{\link{readfict}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{rellife}},
\code{\link{relpersn}},
\code{\link{sprtprsn}},
\code{\link{stateart}},
\code{\link{uniongod}},
\code{\link{volarts}},
\code{\link{wrthless}}

Other Leisure: 
\code{\link{artgn}},
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{buyinfgn}},
\code{\link{camping}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{gamegn}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{hrsrelax}},
\code{\link{huntfish}},
\code{\link{investgn}},
\code{\link{makeart}},
\code{\link{memlit}},
\code{\link{memsport}},
\code{\link{newjobgn}},
\code{\link{partpol}},
\code{\link{peoplegn}},
\code{\link{perform}},
\code{\link{perfrmgn}},
\code{\link{plymusic}},
\code{\link{polcangn}},
\code{\link{poldisgn}},
\code{\link{polinfgn}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{seemovie}},
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timepdwk}},
\code{\link{timerelx}},
\code{\link{travelgn}},
\code{\link{usevcr}},
\code{\link{visart}},
\code{\link{visitart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}},
\code{\link{volarts}}
}
\concept{Leisure}
\concept{Religion and Culture}
\keyword{variable}
