% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_587.R
\name{relate11}
\alias{relate11}
\title{Relation of 11th person (visitor) to head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate11
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
\enumerate{
\item Relationship of eleventh person to head of household
}
}
\section{Values}{

\itemize{
\item \verb{[1]} head of household
\item \verb{[2]} spouse
\item \verb{[3]} child (natural, adopted, step)
\item \verb{[4]} son-/daughter-in-law
\item \verb{[5]} grand/great-grandchild
\item \verb{[6]} parent/parent-in-law
\item \verb{[7]} other relative
\item \verb{[8]} non-relative
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5392/vshow}{GSS Data Explorer page for \code{relate11}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab parent/parent-in-law \tab other relative \tab non-relative \tab son-/daughter-in-law \tab no answer \tab head of household \tab spouse \tab Total \cr
   1975 \tab 1454 \tab 10 \tab 4 \tab 7 \tab 5 \tab 10 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1493 \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1512 \tab 3 \tab 1 \tab 1 \tab 6 \tab 6 \tab 1 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1509 \tab 7 \tab 2 \tab 1 \tab 7 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1532 \cr
   1980 \tab 1436 \tab 6 \tab 2 \tab 6 \tab 6 \tab 6 \tab - \tab 6 \tab - \tab - \tab 1468 \cr
   1982 \tab 1839 \tab 7 \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 6 \tab - \tab - \tab 1860 \cr
   1983 \tab 1575 \tab 3 \tab 2 \tab 3 \tab 10 \tab 6 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1443 \tab 7 \tab - \tab 3 \tab 6 \tab 11 \tab 1 \tab 2 \tab - \tab - \tab 1473 \cr
   1985 \tab 1498 \tab 8 \tab - \tab 10 \tab 8 \tab 7 \tab 1 \tab 2 \tab - \tab - \tab 1534 \cr
   1986 \tab 1452 \tab 5 \tab 2 \tab 3 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab 1470 \cr
   1987 \tab 1794 \tab 6 \tab 3 \tab 6 \tab 3 \tab 6 \tab - \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 1469 \tab 4 \tab - \tab 4 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1522 \tab 1 \tab 3 \tab 3 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 1537 \cr
   1990 \tab 1359 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab - \tab 1372 \cr
   1991 \tab 1498 \tab 4 \tab 2 \tab - \tab 3 \tab 6 \tab - \tab 2 \tab - \tab 2 \tab 1517 \cr
   1993 \tab 1597 \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2970 \tab 8 \tab - \tab 3 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 2881 \tab 2 \tab 1 \tab 2 \tab 2 \tab 7 \tab - \tab 5 \tab - \tab 4 \tab 2904 \cr
   1998 \tab 2805 \tab 11 \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 6 \tab 2832 \cr
   2000 \tab 2808 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2759 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2797 \tab 4 \tab 1 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4462 \tab 13 \tab 1 \tab 7 \tab 11 \tab 15 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2004 \tab 3 \tab 2 \tab 3 \tab 6 \tab 5 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2016 \tab 8 \tab 1 \tab 3 \tab 4 \tab 10 \tab - \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1955 \tab 2 \tab - \tab 4 \tab 6 \tab 7 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2520 \tab 2 \tab 1 \tab 2 \tab 5 \tab 8 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2839 \tab 5 \tab - \tab 5 \tab 8 \tab 10 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2315 \tab 5 \tab 1 \tab 6 \tab 5 \tab 15 \tab - \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 67157 \tab 146 \tab 31 \tab 99 \tab 125 \tab 177 \tab 6 \tab 35 \tab 1 \tab 12 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away10}},
\code{\link{away9}},
\code{\link{gender10}},
\code{\link{mar10}},
\code{\link{mar9}},
\code{\link{old10}},
\code{\link{relate10}},
\code{\link{where10}},
\code{\link{where9}}
}
\concept{Household Members and Structure}
\keyword{variable}
