% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_589.R
\name{relate14}
\alias{relate14}
\title{Relation of 14th person (visitor) to head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate14
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
\enumerate{
\item Relationship of fourteenth person to head of household
}
}
\section{Values}{

\itemize{
\item \verb{[1]} head of household
\item \verb{[2]} spouse
\item \verb{[3]} child (natural, adopted, step)
\item \verb{[4]} son-/daughter-in-law
\item \verb{[5]} grand/great-grandchild
\item \verb{[6]} parent/parent-in-law
\item \verb{[7]} other relative
\item \verb{[8]} non-relative
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5410/vshow}{GSS Data Explorer page for \code{relate14}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab grand/great-grandchild \tab other relative \tab no answer \tab child (natural, adopted, step) \tab Total \cr
   1975 \tab 1487 \tab 2 \tab 1 \tab - \tab - \tab 1490 \cr
   1976 \tab 1498 \tab - \tab 1 \tab - \tab - \tab 1499 \cr
   1977 \tab 1529 \tab 1 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1530 \tab 2 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1466 \tab 1 \tab - \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1598 \tab - \tab - \tab - \tab 1 \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1533 \tab - \tab 1 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1817 \tab 2 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2991 \tab 1 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2830 \tab 1 \tab - \tab - \tab 1 \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4509 \tab - \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2042 \tab - \tab - \tab 2 \tab - \tab 2044 \cr
   2012 \tab 1973 \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2866 \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 67768 \tab 12 \tab 4 \tab 3 \tab 2 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away13}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{mar13}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{relate13}},
\code{\link{where12}},
\code{\link{where13}}
}
\concept{Household Members and Structure}
\keyword{variable}
