% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_586.R
\name{relate8}
\alias{relate8}
\title{Relationship of 8th person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate8
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
\enumerate{
\item Relationship of eighth person to head of household
}
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5374/vshow}{GSS Data Explorer page for \code{relate8}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab parent/parent-in-law \tab non-relative \tab no answer \tab other relative \tab spouse \tab son-/daughter-in-law \tab Total \cr
   1975 \tab 1465 \tab 20 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1490 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1505 \tab 21 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1514 \tab 17 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1532 \cr
   1980 \tab 1454 \tab 11 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1835 \tab 11 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 1584 \tab 12 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab 1465 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1519 \tab 12 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1456 \tab 11 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1804 \tab 7 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 1475 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1520 \tab 8 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1537 \cr
   1990 \tab 1368 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1510 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1602 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab 2985 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2891 \tab 7 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2904 \cr
   1998 \tab 2817 \tab 6 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 2802 \tab 8 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab 2817 \cr
   2002 \tab 2757 \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 2803 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 4495 \tab 12 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 2015 \tab 4 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 2026 \tab 11 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 1967 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab 2531 \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2850 \tab 13 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 2338 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 67419 \tab 246 \tab 64 \tab 8 \tab 13 \tab 9 \tab 28 \tab 1 \tab 1 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away8}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{mar8}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{relate9}},
\code{\link{where7}},
\code{\link{where8}}
}
\concept{Household Members and Structure}
\keyword{variable}
