% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_586.R
\name{relate9}
\alias{relate9}
\title{Relationship of 9th person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate9
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
\enumerate{
\item Relationship of ninth person to head of household
}
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5380/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab no answer \tab other relative \tab spouse \tab son-/daughter-in-law \tab non-relative \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1478 \tab 10 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1494 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1516 \tab 12 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1524 \tab 7 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1460 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1850 \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1592 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1468 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1527 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1461 \tab 6 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1810 \tab 7 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1478 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1528 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1513 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1604 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2991 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2899 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2826 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2804 \tab 6 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 2761 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2765 \cr
   2004 \tab 2808 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4505 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2021 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2033 \tab 4 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 1973 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2531 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2538 \cr
   2016 \tab 2858 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2342 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2348 \cr
   2021 \tab 4025 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 4032 \cr
   2022 \tab 3538 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 72191 \tab 125 \tab 37 \tab 8 \tab 18 \tab 1 \tab 4 \tab 6 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away8}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{mar8}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{relate8}},
\code{\link{where7}},
\code{\link{where8}}

Other Cohabitation: 
\code{\link{relate8}}

Other Family: 
\code{\link{away8}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{mar8}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{relate8}},
\code{\link{where7}},
\code{\link{where8}}

Other Household: 
\code{\link{away8}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{mar8}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{relate8}},
\code{\link{where7}},
\code{\link{where8}}
}
\concept{Cohabitation}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
