% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_373.R
\name{relbstsc}
\alias{relbstsc}
\title{Best policy on stem cell vs narrow interest: relig leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
relbstsc
}
\details{
Question Please look at Card B9. When making policy recommendations about stem cell research, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career
B. Religious leaders
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3374/vshow}{GSS Data Explorer page for \code{relbstsc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 112 \tab 108 \tab 218 \tab 190 \tab 245 \tab 51 \tab 4 \tab 4510 \cr
   2010 \tab 1569 \tab 38 \tab 46 \tab 112 \tab 120 \tab 129 \tab 24 \tab 6 \tab 2044 \cr
   Total \tab 5151 \tab 150 \tab 154 \tab 330 \tab 310 \tab 374 \tab 75 \tab 10 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{medagrsc}},
\code{\link{medbstsc}},
\code{\link{medinfsc}},
\code{\link{polbstsc}},
\code{\link{polinfsc}},
\code{\link{relinfsc}},
\code{\link{scresrch}},
\code{\link{txbiz}},
\code{\link{txeco}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
