% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{relcall}
\alias{relcall}
\title{Contact with most contacted rel}
\source{
General Social Survey https://gss.norc.org
}
\description{
relcall
}
\details{
Question Thinking of all the adult relatives, which one do you have the most contact with?
C. And how often do you have any other contact with your relative besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several x week
\item \verb{[3]} once a week
\item \verb{[4]} once a month
\item \verb{[5]} several x year
\item \verb{[6]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4161/vshow}{GSS Data Explorer page for \code{relcall}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab no answer \tab Total \cr
   1986 \tab 172 \tab 134 \tab 189 \tab 271 \tab 339 \tab 191 \tab 166 \tab 8 \tab 1470 \cr
   Total \tab 172 \tab 134 \tab 189 \tab 271 \tab 339 \tab 191 \tab 166 \tab 8 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
