% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{relcand}
\alias{relcand}
\title{R accepts a person from a different religion or view from r's being candidate of}
\source{
General Social Survey https://gss.norc.org
}
\description{
relcand
}
\details{
Question People belong to different religions and have different religious views.
B. Would you accept a person from a different religion or with a very different religious view from yours being a candidate of the political party you prefer?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely accept
\item \verb{[2]} probably accept
\item \verb{[3]} probably not accept
\item \verb{[4]} definitely not accept
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4702/vshow}{GSS Data Explorer page for \code{relcand}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab definitely accept \tab probably accept \tab probably not accept \tab definitely not accept \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 658 \tab 466 \tab 640 \tab 168 \tab 39 \tab 40 \tab 12 \tab - \tab 2023 \cr
   Total \tab 658 \tab 466 \tab 640 \tab 168 \tab 39 \tab 40 \tab 12 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{befair}},
\code{\link{bible1}},
\code{\link{cantrust}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conbiz}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{hapunhap}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{madenkid}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}

Other Political: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{anomia7}},
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{bribe}},
\code{\link{buspow}},
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf11}},
\code{\link{caninf12}},
\code{\link{caninf13}},
\code{\link{caninf14}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf6}},
\code{\link{caninf7}},
\code{\link{caninf8}},
\code{\link{caninf9}},
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{civic}},
\code{\link{cntctgov}},
\code{\link{congeth}},
\code{\link{contemp}},
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{corruptn}},
\code{\link{demworks}},
\code{\link{discpol}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{engballt}},
\code{\link{ethorgs}},
\code{\link{fixmistk}},
\code{\link{gavepol}},
\code{\link{giveoth}},
\code{\link{govdook}},
\code{\link{governor}},
\code{\link{govngos}},
\code{\link{govtpow}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{interpol}},
\code{\link{intpol}},
\code{\link{joindem}},
\code{\link{knowpols}},
\code{\link{laborpow}},
\code{\link{leftrght}},
\code{\link{libtemp}},
\code{\link{localgvt}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{ntcitvte}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{owneth}},
\code{\link{partpol}},
\code{\link{partyid}},
\code{\link{polactve}},
\code{\link{polcamp}},
\code{\link{poleff1}},
\code{\link{poleff10}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff2}},
\code{\link{poleff20}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}},
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy3}},
\code{\link{polfrel}},
\code{\link{polfunds}},
\code{\link{polgreed}},
\code{\link{polinfo}},
\code{\link{polint}},
\code{\link{polint1}},
\code{\link{polinter}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polnews}},
\code{\link{polpapr}},
\code{\link{polrally}},
\code{\link{polsfair}},
\code{\link{poltv}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{polwww}},
\code{\link{powerun}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prespop}},
\code{\link{refrndms}},
\code{\link{schlhead}},
\code{\link{servepeo}},
\code{\link{signdpet}},
\code{\link{swayvote}},
\code{\link{unrghts}},
\code{\link{usemedia}},
\code{\link{usrep}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{workpol}}

Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{madenkid}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numbwfrd}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{othsch}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}

Other Values: 
\code{\link{agape1}},
\code{\link{agape2}},
\code{\link{agape3}},
\code{\link{agape4}},
\code{\link{impcultr}},
\code{\link{impfinan}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{openmind}},
\code{\link{relmarry}},
\code{\link{united}}
}
\concept{ISSP Religion}
\concept{Political}
\concept{Religion}
\concept{Values}
\keyword{variable}
