% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{relcmfrt}
\alias{relcmfrt}
\title{Find strength and comfort in religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relcmfrt
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
B. I find strength and comfort in my religion.
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never/almst nevr
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2109/vshow}{GSS Data Explorer page for \code{relcmfrt}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never/almst nevr \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 226 \tab 362 \tab 243 \tab 193 \tab 179 \tab 193 \tab 19 \tab 30 \tab 2832 \cr
   Total \tab 1387 \tab 226 \tab 362 \tab 243 \tab 193 \tab 179 \tab 193 \tab 19 \tab 30 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
