% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{relexp}
\alias{relexp}
\title{Have religious experience changed life}
\source{
General Social Survey https://gss.norc.org
}
\description{
relexp
}
\details{
Question Did you ever have a religious or spiritual experience that changed your life?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2021 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2122/vshow}{GSS Data Explorer page for \code{relexp}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{beausprt}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{drama}},
\code{\link{effort}},
\code{\link{feelgod}},
\code{\link{fidgety}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godlove}},
\code{\link{godwatch}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{lesspain}},
\code{\link{localart}},
\code{\link{meditate}},
\code{\link{natarts}},
\code{\link{natlart}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{popmusic}},
\code{\link{privpray}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{relcmfrt}},
\code{\link{rellife}},
\code{\link{relpersn}},
\code{\link{sprtprsn}},
\code{\link{stateart}},
\code{\link{uniongod}},
\code{\link{volarts}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
