% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{relexper}
\alias{relexper}
\title{Turning point in life for religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relexper
}
\details{
Question Has there ever been a turning point in your life when you made a new and personal commitment to religion?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4664/vshow}{GSS Data Explorer page for \code{relexper}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{religkid}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\keyword{variable}
