% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{relext1}
\alias{relext1}
\title{Religious extremists should be allowed to hold public meeting to express views}
\source{
General Social Survey https://gss.norc.org
}
\description{
relext1
}
\details{
Question There are some people whose views are considered extreme by the majority. Consider religious extremists, that is people who believe that their religion is the only true faith and all other religions should be considered as enemies.
A. Do you think such people should be allowed to hold a public meeting to express their views? Would you say…
}
\section{Values}{

\itemize{
\item \code{1} definitely
\item \code{2} probably
\item \code{3} probably not
\item \code{4} definitely not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4703/vshow}{GSS Data Explorer page for \code{relext1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely \tab probably \tab probably not \tab definitely not \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 367 \tab 422 \tab 266 \tab 279 \tab 21 \tab 10 \tab 2023 \cr
   2018 \tab 1173 \tab 260 \tab 391 \tab 291 \tab 204 \tab 19 \tab 10 \tab 2348 \cr
   Total \tab 1831 \tab 627 \tab 813 \tab 557 \tab 483 \tab 40 \tab 20 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{ancestrs}},
\code{\link{mywaygod}},
\code{\link{nirvana}},
\code{\link{reincar}},
\code{\link{relcand}},
\code{\link{relext2}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\keyword{variable}
