% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{relext2}
\alias{relext2}
\title{People should be allowed to publish books expressing their views}
\source{
General Social Survey https://gss.norc.org
}
\description{
relext2
}
\details{
Question There are some people whose views are considered extreme by the majority. Consider religious extremists, that is people who believe that their religion is the only true faith and all other religions should be considered as enemies.
B. Do you think such people should be allowed to publish books expressing their views? Would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely
\item \verb{[2]} probably
\item \verb{[3]} probably not
\item \verb{[4]} definitely not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4704/vshow}{GSS Data Explorer page for \code{relext2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely \tab probably \tab probably not \tab definitely not \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 538 \tab 538 \tab 143 \tab 117 \tab 21 \tab 8 \tab 2023 \cr
   Total \tab 658 \tab 538 \tab 538 \tab 143 \tab 117 \tab 21 \tab 8 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{ancestrs}},
\code{\link{mywaygod}},
\code{\link{nirvana}},
\code{\link{reincar}},
\code{\link{relcand}},
\code{\link{relext1}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\keyword{variable}
