% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{relgeneq}
\alias{relgeneq}
\title{R's religion treats men and women equally}
\source{
General Social Survey https://gss.norc.org
}
\description{
relgeneq
}
\details{
Question In your opinion, does your religion treat men and women equally, treat men better than women, or treat women better than men?
}
\section{Values}{

\itemize{
\item \code{1} i don't belong to or follow any religion
\item \code{2} treats men and women equally
\item \code{3} treats men better than women
\item \code{4} treats women better than men
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7280/vshow}{GSS Data Explorer page for \code{relgeneq}}.
}

\keyword{variable}
