% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{relgrade}
\alias{relgrade}
\title{Grade given religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relgrade
}
\details{
Question Using the school grading system of A,B,C,D, and F for Fail, how Do you rate your church or religious group in meeting your spiritual needs?
}
\section{Values}{

\itemize{
\item \code{1} a
\item \code{2} b
\item \code{3} c
\item \code{4} d
\item \code{5} f
\item \code{6} no religion
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/381/vshow}{GSS Data Explorer page for \code{relgrade}}.
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}}

Other Religion: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
