% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{relhh1}
\alias{relhh1}
\title{Relationship of person 1 to head of household}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhh1
}
\details{
Question What is (PERSON)'s relationship to head of household?
A. Relationship of first person to head of household?
}
\section{Values}{

\itemize{
\item \verb{[1]} head of household
\item \verb{[2]} spouse
\item \verb{[3]} partner, fiance(e), boyfriend, girlfriend, etc.
\item \verb{[4]} child, not specified
\item \verb{[5]} grandchild, great-grandchild
\item \verb{[6]} son-in-law, daughter-in-law
\item \verb{[7]} parent
\item \verb{[8]} parent-in-law
\item \verb{[9]} grandparent, great-grandparent
\item \verb{[10]} uncle or aunt
\item \verb{[11]} niece or nephew
\item \verb{[12]} cousin
\item \verb{[13]} brother, sister (full, half, or step sibling)
\item \verb{[14]} sibling-in-law
\item \verb{[15]} other relative (great-aunt, grandniece, etc.)
\item \verb{[16]} roommate, housemate
\item \verb{[17]} friend
\item \verb{[18]} child of non-relative
\item \verb{[19]} other non-relative (employee, boarder)
\item \verb{[41]} natural/biological child
\item \verb{[42]} adopted child
\item \verb{[43]} stepchild
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5439/vshow}{GSS Data Explorer page for \code{relhh1}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab head of household \tab not available in this release \tab Total \cr
   2000 \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 0 \tab 26698 \tab 4032 \tab 30730 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{famgen}},
\code{\link{hhtype}},
\code{\link{hhtype1}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}}
}
\concept{Household Members and Structure}
\keyword{variable}
