% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{relhh3}
\alias{relhh3}
\title{Relationship of person 3 to head of household}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhh3
}
\details{
Question What is (PERSON)’s relationship to head of household?
C. Relationship of third person to head of household?
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, not specified
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{41} natural/biological child
\item \code{42} adopted child
\item \code{43} stepchild
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5441/vshow}{GSS Data Explorer page for \code{relhh3}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab child, not specified \tab grandchild, great-grandchild \tab son-in-law, daughter-in-law \tab parent \tab parent-in-law \tab grandparent, great-grandparent \tab uncle or aunt \tab niece or nephew \tab cousin \tab brother, sister (full, half, or step sibling) \tab sibling-in-law \tab other relative (great-aunt, grandniece, etc.) \tab roommate, housemate \tab friend \tab child of non-relative \tab other non-relative (employee, boarder) \tab natural/biological child \tab adopted child \tab stepchild \tab don't know \tab no answer \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab not available in this release \tab Total \cr
   2000 \tab 1668 \tab 137 \tab 33 \tab 10 \tab 20 \tab 9 \tab 4 \tab 1 \tab 7 \tab 1 \tab 15 \tab 7 \tab 3 \tab 27 \tab 7 \tab 13 \tab 9 \tab 723 \tab 16 \tab 48 \tab 1 \tab 58 \tab - \tab - \tab 2817 \cr
   2002 \tab 1815 \tab 193 \tab 39 \tab 15 \tab 5 \tab 8 \tab 3 \tab 2 \tab 7 \tab 2 \tab 19 \tab 4 \tab 3 \tab 25 \tab 12 \tab 14 \tab 12 \tab 524 \tab 10 \tab 47 \tab 1 \tab 5 \tab - \tab - \tab 2765 \cr
   2004 \tab 1780 \tab 308 \tab 34 \tab 11 \tab 18 \tab 6 \tab 1 \tab 1 \tab 7 \tab 3 \tab 13 \tab - \tab 4 \tab 15 \tab 4 \tab 5 \tab 5 \tab 546 \tab 11 \tab 34 \tab 3 \tab 1 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 2778 \tab 546 \tab 44 \tab 9 \tab 28 \tab 17 \tab 2 \tab 2 \tab 13 \tab 2 \tab 19 \tab 9 \tab 5 \tab 44 \tab 16 \tab 7 \tab 11 \tab 878 \tab 12 \tab 59 \tab 2 \tab 1 \tab 6 \tab - \tab 4510 \cr
   2008 \tab 1221 \tab 350 \tab 34 \tab 1 \tab 10 \tab 8 \tab 1 \tab 2 \tab 9 \tab 4 \tab 9 \tab 6 \tab 2 \tab 16 \tab 5 \tab 5 \tab 2 \tab 304 \tab 2 \tab 28 \tab 1 \tab 2 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1319 \tab 371 \tab 26 \tab 10 \tab 11 \tab 4 \tab 1 \tab 2 \tab 4 \tab 2 \tab 9 \tab 7 \tab 1 \tab 24 \tab 9 \tab 8 \tab 2 \tab 212 \tab 4 \tab 13 \tab - \tab 3 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 1259 \tab 325 \tab 14 \tab 3 \tab 19 \tab 1 \tab - \tab 12 \tab 8 \tab 2 \tab 10 \tab 12 \tab 3 \tab 21 \tab 9 \tab 8 \tab 2 \tab 226 \tab 5 \tab 27 \tab 2 \tab 5 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 1642 \tab 434 \tab 21 \tab 8 \tab 16 \tab 6 \tab 4 \tab 1 \tab 4 \tab 4 \tab 13 \tab 3 \tab - \tab 38 \tab 11 \tab 8 \tab 5 \tab 286 \tab 6 \tab 17 \tab 4 \tab 1 \tab 6 \tab - \tab 2538 \cr
   2016 \tab 1919 \tab 580 \tab 37 \tab 14 \tab 18 \tab 15 \tab - \tab 1 \tab 14 \tab 3 \tab 7 \tab 9 \tab 3 \tab 37 \tab 10 \tab 5 \tab 6 \tab 150 \tab 6 \tab 29 \tab 2 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 1561 \tab 402 \tab 28 \tab 7 \tab 19 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2 \tab 12 \tab 7 \tab 4 \tab 34 \tab 2 \tab 4 \tab 5 \tab 219 \tab 4 \tab 26 \tab - \tab 1 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 16962 \tab 3646 \tab 310 \tab 88 \tab 164 \tab 78 \tab 17 \tab 26 \tab 76 \tab 25 \tab 126 \tab 64 \tab 28 \tab 281 \tab 85 \tab 77 \tab 59 \tab 4068 \tab 76 \tab 328 \tab 16 \tab 78 \tab 20 \tab 4032 \tab 30730 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{famgen}},
\code{\link{hhtype}},
\code{\link{hhtype1}},
\code{\link{relhh1}},
\code{\link{relhh2}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}}
}
\concept{Household Members and Structure}
\keyword{variable}
