% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_593.R
\name{relhh7}
\alias{relhh7}
\title{Relationship of person 7 to head of household}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhh7
}
\details{
Question What is (PERSON)’s relationship to head of household?
G. Relationship of seventh person to head of household?
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, not specified
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{41} natural/biological child
\item \code{42} adopted child
\item \code{43} stepchild
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5445/vshow}{GSS Data Explorer page for \code{relhh7}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab child, not specified \tab grandchild, great-grandchild \tab parent \tab niece or nephew \tab cousin \tab sibling-in-law \tab natural/biological child \tab adopted child \tab stepchild \tab no answer \tab friend \tab child of non-relative \tab parent-in-law \tab other relative (great-aunt, grandniece, etc.) \tab roommate, housemate \tab brother, sister (full, half, or step sibling) \tab other non-relative (employee, boarder) \tab son-in-law, daughter-in-law \tab not available in this release \tab Total \cr
   2000 \tab 2786 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 16 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2741 \tab 2 \tab 7 \tab - \tab 1 \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2786 \tab 7 \tab 3 \tab - \tab 1 \tab - \tab - \tab 8 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4457 \tab 15 \tab 7 \tab - \tab 1 \tab - \tab - \tab 22 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2002 \tab 7 \tab 6 \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2003 \tab 12 \tab 10 \tab - \tab 5 \tab - \tab - \tab 10 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1951 \tab 7 \tab 6 \tab 1 \tab 2 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab 2515 \tab 10 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2538 \cr
   2016 \tab 2830 \tab 18 \tab 7 \tab - \tab 1 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2327 \tab 5 \tab 6 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 26398 \tab 85 \tab 58 \tab 2 \tab 14 \tab 2 \tab 2 \tab 94 \tab 7 \tab 9 \tab 5 \tab 3 \tab 3 \tab 4 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 4032 \tab 30730 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh13}},
\code{\link{relhh14}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh8}},
\code{\link{relhh9}}
}
\concept{Household Members and Structure}
\keyword{variable}
