% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{relhhd10}
\alias{relhhd10}
\title{Relation of 1oth person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhhd10
}
\details{
Question What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?  SEE CODES BELOW ROSTER
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, natural or adopted, stepchild
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5425/vshow}{GSS Data Explorer page for \code{relhhd10}}.
}

\seealso{
Other Household Members and Structure: 
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{respnum}}

Other Cohabitation: 
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}}

Other Family: 
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}}

Other Household: 
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{respnum}}
}
\concept{Cohabitation}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
