% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{relhhd11}
\alias{relhhd11}
\title{Relation of 11th person (visitor) to head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhhd11
}
\details{
Question What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?  SEE CODES BELOW ROSTER
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, natural or adopted, stepchild
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5426/vshow}{GSS Data Explorer page for \code{relhhd11}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab child, natural or adopted, stepchild \tab parent \tab parent-in-law \tab sibling-in-law \tab friend \tab other non-relative (employee, boarder) \tab niece or nephew \tab brother, sister (full, half, or step sibling) \tab no answer \tab spouse \tab grandchild, great-grandchild \tab roommate, housemate \tab child of non-relative \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab son-in-law, daughter-in-law \tab uncle or aunt \tab cousin \tab other relative (great-aunt, grandniece, etc.) \tab not available in this release \tab Total \cr
   1993 \tab 1597 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2970 \tab 8 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2881 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2805 \tab 11 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2808 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2759 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2797 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4462 \tab 13 \tab 5 \tab 2 \tab 3 \tab 10 \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 2004 \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2023 \cr
   2010 \tab 2016 \tab 8 \tab 1 \tab 2 \tab - \tab 6 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1955 \tab 2 \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab 2520 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2839 \tab 5 \tab 4 \tab 1 \tab 1 \tab 8 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab 2315 \tab 5 \tab 4 \tab 2 \tab 1 \tab 8 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 40272 \tab 71 \tab 26 \tab 21 \tab 10 \tab 54 \tab 17 \tab 15 \tab 22 \tab 10 \tab 10 \tab 10 \tab 13 \tab 2 \tab 9 \tab 2 \tab 3 \tab 7 \tab 2 \tab 4032 \tab 44608 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{relhhd10}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{respnum}}
}
\concept{Household Members and Structure}
\keyword{variable}
