% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{relhhd13}
\alias{relhhd13}
\title{Relation of 13th person (visitor) to head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhhd13
}
\details{
Question What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?  SEE CODES BELOW ROSTER
}
\section{Values}{

\itemize{
\item \verb{[1]} head of household
\item \verb{[2]} spouse
\item \verb{[3]} partner, fiance(e), boyfriend, girlfriend, etc.
\item \verb{[4]} child, natural or adopted, stepchild
\item \verb{[5]} grandchild, great-grandchild
\item \verb{[6]} son-in-law, daughter-in-law
\item \verb{[7]} parent
\item \verb{[8]} parent-in-law
\item \verb{[9]} grandparent, great-grandparent
\item \verb{[10]} uncle or aunt
\item \verb{[11]} niece or nephew
\item \verb{[12]} cousin
\item \verb{[13]} brother, sister (full, half, or step sibling)
\item \verb{[14]} sibling-in-law
\item \verb{[15]} other relative (great-aunt, grandniece, etc.)
\item \verb{[16]} roommate, housemate
\item \verb{[17]} friend
\item \verb{[18]} child of non-relative
\item \verb{[19]} other non-relative (employee, boarder)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5428/vshow}{GSS Data Explorer page for \code{relhhd13}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab grandchild, great-grandchild \tab child, natural or adopted, stepchild \tab child of non-relative \tab no answer \tab niece or nephew \tab don't know \tab not available in this release \tab Total \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2991 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2903 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2826 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2816 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4508 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 2022 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 2042 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1973 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2536 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2865 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 40557 \tab 6 \tab 5 \tab 2 \tab 3 \tab 2 \tab 1 \tab 4032 \tab 44608 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{respnum}}
}
\concept{Household Members and Structure}
\keyword{variable}
