% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{relhrs2}
\alias{relhrs2}
\title{Hours religious act outside of home}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhrs2
}
\details{
Question B. In the past month, about how many hours have you spent doing religious services activities outside your home (such as attending religious services, prayer groups, Bible studies, fellowship meetings, church leadership meetings, etc.)?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1144/vshow}{GSS Data Explorer page for \code{relhrs2}}.
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{givearts}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{kid5up}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}}
}
\concept{Religion}
\keyword{variable}
