% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{relid1}
\alias{relid1}
\title{Protestant identification 1st mention}
\source{
General Social Survey https://gss.norc.org
}
\description{
relid1
}
\details{
Question B. When it comes to your religious identity, would you say you are a pentecostal, fundamentalist, evangelical, mainline, or liberal Protestant, or do none of these describe you?
}
\section{Values}{

\itemize{
\item \verb{[1]} pentecostal
\item \verb{[2]} fundamentalist
\item \verb{[3]} evangelical
\item \verb{[4]} mainline
\item \verb{[5]} liberal
\item \verb{[6]} none
\item \verb{[7]} other (specify)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5326/vshow}{GSS Data Explorer page for \code{relid1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab pentecostal \tab fundamentalist \tab evangelical \tab mainline \tab liberal \tab none \tab other (specify) \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1367 \tab 128 \tab 74 \tab 135 \tab 170 \tab 219 \tab 641 \tab 38 \tab 11 \tab 34 \tab 2817 \cr
   Total \tab 1367 \tab 128 \tab 74 \tab 135 \tab 170 \tab 219 \tab 641 \tab 38 \tab 11 \tab 34 \tab 2817 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
