% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{relid2}
\alias{relid2}
\title{Protestant identification 2nd mention}
\source{
General Social Survey https://gss.norc.org
}
\description{
relid2
}
\details{
Question C. Do you consider yourself to be a member of (READ REMAINING GROUPS)?
}
\section{Values}{

\itemize{
\item \verb{[1]} pentecostal
\item \verb{[2]} fundamentalist
\item \verb{[3]} evangelical
\item \verb{[4]} mainline
\item \verb{[5]} liberal
\item \verb{[6]} none
\item \verb{[7]} other (specify)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5327/vshow}{GSS Data Explorer page for \code{relid2}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab pentecostal \tab fundamentalist \tab evangelical \tab mainline \tab liberal \tab none \tab other (specify) \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2008 \tab 5 \tab 11 \tab 19 \tab 17 \tab 10 \tab 480 \tab 3 \tab 11 \tab 253 \tab 2817 \cr
   Total \tab 2008 \tab 5 \tab 11 \tab 19 \tab 17 \tab 10 \tab 480 \tab 3 \tab 11 \tab 253 \tab 2817 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
