% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{relidbst}
\alias{relidbst}
\title{Protestant identification best group}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidbst
}
\details{
Question E. Which of these best describes what you are?
}
\section{Values}{

\itemize{
\item \verb{[1]} pentecostal
\item \verb{[2]} fundamentalist
\item \verb{[3]} evangelical
\item \verb{[4]} mainline
\item \verb{[5]} liberal
\item \verb{[6]} none
\item \verb{[7]} other (specify)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5329/vshow}{GSS Data Explorer page for \code{relidbst}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab pentecostal \tab fundamentalist \tab evangelical \tab mainline \tab liberal \tab none \tab other (specify) \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2488 \tab 9 \tab 10 \tab 7 \tab 9 \tab 6 \tab 2 \tab 1 \tab 11 \tab 274 \tab 2817 \cr
   Total \tab 2488 \tab 9 \tab 10 \tab 7 \tab 9 \tab 6 \tab 2 \tab 1 \tab 11 \tab 274 \tab 2817 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
