% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{relidwe}
\alias{relidwe}
\title{How often r says 'we' when talking about r's religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidwe
}
\details{
Question When talking about your religion, how often do you say we instead of they?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} rarely
\item \verb{[3]} some of the time
\item \verb{[4]} most of the time
\item \verb{[5]} all of the time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7528/vshow}{GSS Data Explorer page for \code{relidwe}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab never \tab rarely \tab some of the time \tab most of the time \tab all of the time \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 393 \tab 785 \tab 586 \tab 857 \tab 735 \tab 469 \tab 21 \tab 6 \tab 82 \tab 98 \tab 4032 \cr
   Total \tab 393 \tab 785 \tab 586 \tab 857 \tab 735 \tab 469 \tab 21 \tab 6 \tab 82 \tab 98 \tab 4032 \cr
}
}

\keyword{variable}
