% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{reliedon}
\alias{reliedon}
\title{Ppl at work can be relied on when needs help}
\source{
General Social Survey https://gss.norc.org
}
\description{
reliedon
}
\details{
Question K. People at work with can be relied on when someone needs help.
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not too true
\item \verb{[4]} not at all true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3008/vshow}{GSS Data Explorer page for \code{reliedon}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 1104 \tab 605 \tab 58 \tab 38 \tab 20 \tab 42 \tab 2812 \cr
   Total \tab 945 \tab 1104 \tab 605 \tab 58 \tab 38 \tab 20 \tab 42 \tab 2812 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
