% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_104.R
\name{relig3}
\alias{relig3}
\title{Religious preference of person #3}
\source{
General Social Survey https://gss.norc.org
}
\description{
relig3
}
\details{
Question What is (NAME 1)'s religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?  PROBE:
What is your best guess?
ASK FOR EACH NAME REPEATING CATEGORIES AS NECESSARY.
3. Religion of Third Person
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/952/vshow}{GSS Data Explorer page for \code{relig3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1985 \tab 599 \tab 512 \tab 241 \tab 26 \tab 56 \tab 20 \tab 73 \tab 7 \tab 1534 \cr
   2004 \tab 2240 \tab 242 \tab 120 \tab 20 \tab 49 \tab 35 \tab 58 \tab 48 \tab 2812 \cr
   Total \tab 2839 \tab 754 \tab 361 \tab 46 \tab 105 \tab 55 \tab 131 \tab 55 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{talkpol1}},
\code{\link{talkpol2}}
}
\concept{Social Networks}
\keyword{variable}
