% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_104.R
\name{relig5}
\alias{relig5}
\title{Religious preference of person #5}
\source{
General Social Survey https://gss.norc.org
}
\description{
relig5
}
\details{
Question What is (NAME 1)'s religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?  PROBE:
What is your best guess?
ASK FOR EACH NAME REPEATING CATEGORIES AS NECESSARY.
5. Religion of Fifth Person
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/954/vshow}{GSS Data Explorer page for \code{relig5}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 198 \tab 111 \tab 14 \tab 21 \tab 12 \tab 22 \tab 3 \tab 1534 \cr
   2004 \tab 2600 \tab 71 \tab 42 \tab 5 \tab 18 \tab 8 \tab 22 \tab 46 \tab 2812 \cr
   Total \tab 3753 \tab 269 \tab 153 \tab 19 \tab 39 \tab 20 \tab 44 \tab 49 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{talkpol1}},
\code{\link{talkpol2}}
}
\concept{Social Networks}
\keyword{variable}
