% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{relig_next}
\alias{relig_next}
\title{Religion - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
relig_next
}
\details{
Question What is your religious preference?
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[6]} buddhism
\item \verb{[7]} hinduism
\item \verb{[8]} other eastern religions
\item \verb{[9]} muslim/islam
\item \verb{[10]} orthodox-christian
\item \verb{[11]} christian
\item \verb{[12]} native american
\item \verb{[13]} inter-nondenominational
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8009/vshow}{GSS Data Explorer page for \code{relig_next}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab buddhism \tab hinduism \tab muslim/islam \tab orthodox-christian \tab christian \tab native american \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 362 \tab 242 \tab 41 \tab 384 \tab 146 \tab 10 \tab 7 \tab 12 \tab 5 \tab 21 \tab 1 \tab 2308 \tab 5 \tab 3544 \cr
   Total \tab 0 \tab 362 \tab 242 \tab 41 \tab 384 \tab 146 \tab 10 \tab 7 \tab 12 \tab 5 \tab 21 \tab 1 \tab 2308 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
