% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{religid}
\alias{religid}
\title{R's religious identity}
\source{
General Social Survey https://gss.norc.org
}
\description{
religid
}
\details{
Question A. When it comes to your religious identity, would you say you are a fundamentalist, evangelical, mainline, or liberal Protestant or do none of these describe you?
}
\section{Values}{

\itemize{
\item \verb{[1]} fundamentalist
\item \verb{[2]} evangelical
\item \verb{[3]} mainline
\item \verb{[4]} liberal
\item \verb{[5]} none
\item \verb{[6]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5325/vshow}{GSS Data Explorer page for \code{religid}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab fundamentalist \tab evangelical \tab mainline \tab liberal \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1240 \tab 162 \tab 208 \tab 286 \tab 265 \tab 561 \tab 51 \tab 14 \tab 117 \tab 2904 \cr
   1998 \tab 2019 \tab 74 \tab 95 \tab 118 \tab 122 \tab 232 \tab 34 \tab 17 \tab 121 \tab 2832 \cr
   Total \tab 3259 \tab 236 \tab 303 \tab 404 \tab 387 \tab 793 \tab 85 \tab 31 \tab 238 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
