% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{religimp}
\alias{religimp}
\title{How important is religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
religimp
}
\details{
Question How important is religion in your life -- very important, somewhat important, not too important, or not at all important?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} somewhat important
\item \verb{[3]} not too important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7525/vshow}{GSS Data Explorer page for \code{religimp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab somewhat important \tab not too important \tab not at all important \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 390 \tab 1124 \tab 923 \tab 668 \tab 894 \tab 2 \tab 1 \tab 30 \tab 4032 \cr
   Total \tab 390 \tab 1124 \tab 923 \tab 668 \tab 894 \tab 2 \tab 1 \tab 30 \tab 4032 \cr
}
}

\keyword{variable}
