% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{religinf}
\alias{religinf}
\title{The u.s. would be a better country if religion had less influence}
\source{
General Social Survey https://gss.norc.org
}
\description{
religinf
}
\details{
Question Do you agree or disagree with the following statements?
C. The U.S. would be a better country if religion had less influence.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2021 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4690/vshow}{GSS Data Explorer page for \code{religinf}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   1998 \tab 1548 \tab 51 \tab 131 \tab 321 \tab 479 \tab 209 \tab 49 \tab 44 \tab - \tab - \tab 2832 \cr
   2021 \tab 427 \tab 514 \tab 565 \tab 1069 \tab 890 \tab 521 \tab 3 \tab 5 \tab 38 \tab - \tab 4032 \cr
   2024 \tab 1106 \tab 339 \tab 343 \tab 582 \tab 509 \tab 362 \tab 34 \tab 11 \tab 23 \tab - \tab 3309 \cr
   Total \tab 3081 \tab 904 \tab 1039 \tab 1972 \tab 1878 \tab 1092 \tab 86 \tab 60 \tab 61 \tab 0 \tab 10173 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religint}},
\code{\link{socright}},
\code{\link{trustsci}}

Other Religion: 
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religint}},
\code{\link{socright}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
