% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{religint}
\alias{religint}
\title{Religious people are often too intolerant}
\source{
General Social Survey https://gss.norc.org
}
\description{
religint
}
\details{
Question Do you agree or disagree with the following statements?
B. People with very strong religious beliefs are often too intolerant of others.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4689/vshow}{GSS Data Explorer page for \code{religint}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 149 \tab 433 \tab 272 \tab 269 \tab 66 \tab 53 \tab 42 \tab 2832 \cr
   2008 \tab 658 \tab 283 \tab 627 \tab 199 \tab 200 \tab 32 \tab 21 \tab 3 \tab 2023 \cr
   2018 \tab 1173 \tab 179 \tab 524 \tab 239 \tab 181 \tab 34 \tab 10 \tab 8 \tab 2348 \cr
   Total \tab 3379 \tab 611 \tab 1584 \tab 710 \tab 650 \tab 132 \tab 84 \tab 53 \tab 7203 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{socright}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
