% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{religsp}
\alias{religsp}
\title{R's spouse's religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
religsp
}
\details{
Question What is your husband's/wife's religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} orthodox greek, russian, etc
\item \verb{[5]} moslem
\item \verb{[6]} other -specify
\item \verb{[7]} no religion
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4671/vshow}{GSS Data Explorer page for \code{religsp}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab orthodox greek, russian, etc \tab moslem \tab other -specify \tab no religion \tab don't know \tab no answer \tab Total \cr
   1991 \tab 799 \tab 413 \tab 207 \tab 16 \tab 1 \tab 1 \tab 9 \tab 47 \tab 12 \tab 12 \tab 1517 \cr
   1998 \tab 2221 \tab 331 \tab 137 \tab 17 \tab 6 \tab 3 \tab 9 \tab 59 \tab 21 \tab 28 \tab 2832 \cr
   2008 \tab 1378 \tab 382 \tab 153 \tab 14 \tab 6 \tab 2 \tab 9 \tab 68 \tab 9 \tab 2 \tab 2023 \cr
   Total \tab 4398 \tab 1126 \tab 497 \tab 47 \tab 13 \tab 6 \tab 27 \tab 174 \tab 42 \tab 42 \tab 6372 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{relexper}},
\code{\link{religkid}}
}
\concept{ISSP Religion}
\keyword{variable}
