% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_373.R
\name{relinfsc}
\alias{relinfsc}
\title{Influence on funding for stem cell rsch: religious leaders}
\source{
General Social Survey https://gss.norc.org
}
\description{
relinfsc
}
\details{
Question Please look at Card B8. How much influence should each of the following groups have in deciding about government funding for stem cell research?
B. Religious leaders. (Would you say a great deal of influence, a fair amount, a little influence, or none at all?)
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal of influence
\item \verb{[2]} a fair amount
\item \verb{[3]} a little influence
\item \verb{[4]} none at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3371/vshow}{GSS Data Explorer page for \code{relinfsc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal of influence \tab a fair amount \tab a little influence \tab none at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 68 \tab 197 \tab 345 \tab 265 \tab 49 \tab 4 \tab 4510 \cr
   2010 \tab 1569 \tab 34 \tab 85 \tab 174 \tab 159 \tab 19 \tab 4 \tab 2044 \cr
   Total \tab 5151 \tab 102 \tab 282 \tab 519 \tab 424 \tab 68 \tab 8 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{medagrsc}},
\code{\link{medbstsc}},
\code{\link{medinfsc}},
\code{\link{polbstsc}},
\code{\link{polinfsc}},
\code{\link{relbstsc}},
\code{\link{scresrch}},
\code{\link{txbiz}},
\code{\link{txeco}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
