% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{relitena}
\alias{relitena}
\title{Strength of affiliation}
\source{
General Social Survey https://gss.norc.org
}
\description{
relitena
}
\details{
Question C. Would you call yourself a strong Jew or not a very strong Jew?
}
\section{Values}{

\itemize{
\item \code{1} strong
\item \code{2} not very strong
\item \code{3} somewhat strong
\item \code{4} no religion
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/299/vshow}{GSS Data Explorer page for \code{relitena}}.
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jew}},
\code{\link{jewaj}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{spattend}}
}
\concept{Core}
\keyword{variable}
