% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{rellaw}
\alias{rellaw}
\title{Relationship to the attorney}
\source{
General Social Survey https://gss.norc.org
}
\description{
rellaw
}
\details{
Question What was your relationship to the attorney who provided the services, or to another attorney in the firm? Please indicate what your closest relationship was.
}
\section{Values}{

\itemize{
\item \verb{[1]} a relative including in-laws
\item \verb{[2]} a friend or acquaintance
\item \verb{[3]} a friend of a friend or relative etc
\item \verb{[4]} previous business dealings
\item \verb{[5]} no prior relationship
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1837/vshow}{GSS Data Explorer page for \code{rellaw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a relative including in-laws \tab a friend or acquaintance \tab a friend of a friend or relative etc \tab previous business dealings \tab no prior relationship \tab no answer \tab Total \cr
   1996 \tab 2254 \tab 9 \tab 111 \tab 43 \tab 87 \tab 395 \tab 5 \tab 2904 \cr
   Total \tab 2254 \tab 9 \tab 111 \tab 43 \tab 87 \tab 395 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{satlawyr}}
}
\concept{Markets}
\keyword{variable}
