% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{relmeet}
\alias{relmeet}
\title{Allowed to hold pub meeting for religious extremist}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmeet
}
\details{
Question Should religious extremists be allowed to hold public meetings?
}
\section{Values}{

\itemize{
\item \code{1} should definitely be allowed
\item \code{2} should probably be allowed
\item \code{3} should probably not be allowed
\item \code{4} should definitely not be allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4917/vshow}{GSS Data Explorer page for \code{relmeet}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\concept{Religion}
\keyword{variable}
