% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{relmhsp2}
\alias{relmhsp2}
\title{Patient was immediate family}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmhsp2
}
\details{
Question Was this a relative, a close friend, or just someone you didn't know very well? (CIRCLE ALL THAT APPLY) PROBE UNTIL "NO": ANYONE ELSE?
B. Immediate Family.
}
\section{Values}{

\itemize{
\item \verb{[1]} imm. fam
\item \verb{[2]} not imm. fam
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2060/vshow}{GSS Data Explorer page for \code{relmhsp2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab imm. fam \tab not imm. fam \tab no answer \tab Total \cr
   1996 \tab 2531 \tab 92 \tab 272 \tab 9 \tab 2904 \cr
   Total \tab 2531 \tab 92 \tab 272 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{knwmhosp}},
\code{\link{relmhsp1}}
}
\concept{Mental Health}
\keyword{variable}
