% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{relmhsp5}
\alias{relmhsp5}
\title{Patient was acquaintance}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmhsp5
}
\details{
Question Was this a relative, a close friend, or just someone you didn't know very well? (CIRCLE ALL THAT APPLY) PROBE UNTIL "NO": ANYONE ELSE?
E. Acquaintances.
}
\section{Values}{

\itemize{
\item \verb{[1]} acquaintance
\item \verb{[2]} not acquaintance
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2063/vshow}{GSS Data Explorer page for \code{relmhsp5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab acquaintance \tab not acquaintance \tab no answer \tab Total \cr
   1996 \tab 2531 \tab 88 \tab 276 \tab 9 \tab 2904 \cr
   Total \tab 2531 \tab 88 \tab 276 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo1}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp6}}
}
\concept{Mental Health}
\keyword{variable}
