% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{relmhsp6}
\alias{relmhsp6}
\title{Patient was other}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmhsp6
}
\details{
Question Was this a relative, a close friend, or just someone you didn't know very well? (CIRCLE ALL THAT APPLY) PROBE UNTIL "NO": ANYONE ELSE?
F. Other (specify).
}
\section{Values}{

\itemize{
\item \code{1} other
\item \code{2} not other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2064/vshow}{GSS Data Explorer page for \code{relmhsp6}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab other \tab not other \tab no answer \tab Total \cr
   1996 \tab 2531 \tab 26 \tab 338 \tab 9 \tab 2904 \cr
   Total \tab 2531 \tab 26 \tab 338 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo1}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}}
}
\concept{Mental Health}
\keyword{variable}
