% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{relmost}
\alias{relmost}
\title{Adult relative r has most contact with}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmost
}
\details{
Question Thinking of all the adult relatives, which one do you have the most contact with?
}
\section{Values}{

\itemize{
\item \verb{[1]} grmother
\item \verb{[2]} grfather
\item \verb{[3]} grdaughter
\item \verb{[4]} grandson
\item \verb{[5]} aunt
\item \verb{[6]} uncle
\item \verb{[7]} mother-in-law
\item \verb{[8]} father-in-law
\item \verb{[9]} sister-in-law
\item \verb{[10]} brother-in-law
\item \verb{[11]} female relative
\item \verb{[12]} male relative
\item \verb{[13]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4158/vshow}{GSS Data Explorer page for \code{relmost}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab grmother \tab grfather \tab grdaughter \tab grandson \tab aunt \tab uncle \tab mother-in-law \tab father-in-law \tab sister-in-law \tab brother-in-law \tab female relative \tab male relative \tab none \tab no answer \tab Total \cr
   1986 \tab - \tab 110 \tab 19 \tab 68 \tab 44 \tab 149 \tab 54 \tab 213 \tab 47 \tab 198 \tab 137 \tab 191 \tab 100 \tab 134 \tab 6 \tab 1470 \cr
   Total \tab 0 \tab 110 \tab 19 \tab 68 \tab 44 \tab 149 \tab 54 \tab 213 \tab 47 \tab 198 \tab 137 \tab 191 \tab 100 \tab 134 \tab 6 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
