% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_366.R
\name{relneg}
\alias{relneg}
\title{Any turning point when less committed to religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relneg
}
\details{
Question Has there ever been a turning point in your life when you became less commited to religion?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3313/vshow}{GSS Data Explorer page for \code{relneg}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3041 \tab 568 \tab 889 \tab 8 \tab 4 \tab 4510 \cr
   2008 \tab - \tab 767 \tab 1241 \tab 9 \tab 6 \tab 2023 \cr
   2010 \tab - \tab 751 \tab 1275 \tab 11 \tab 7 \tab 2044 \cr
   Total \tab 3041 \tab 2086 \tab 3405 \tab 28 \tab 17 \tab 8577 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
