% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{relpast}
\alias{relpast}
\title{In america, religion represents the past and not the future.}
\source{
General Social Survey https://gss.norc.org
}
\description{
relpast
}
\details{
Question In America, religion represents the past and not the future.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7284/vshow}{GSS Data Explorer page for \code{relpast}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2018 \tab 1173 \tab 33 \tab 184 \tab 251 \tab 531 \tab 140 \tab 22 \tab 14 \tab 2348 \cr
   Total \tab 1173 \tab 33 \tab 184 \tab 251 \tab 531 \tab 140 \tab 22 \tab 14 \tab 2348 \cr
}
}

\keyword{variable}
