% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{relrlvnt}
\alias{relrlvnt}
\title{In america, religion is just as relevant to life today as it was in the past.}
\source{
General Social Survey https://gss.norc.org
}
\description{
relrlvnt
}
\details{
Question In America, religion is just as relevant to life today as it was in the past.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7285/vshow}{GSS Data Explorer page for \code{relrlvnt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2018 \tab 1173 \tab 147 \tab 443 \tab 159 \tab 346 \tab 49 \tab 19 \tab 12 \tab 2348 \cr
   Total \tab 1173 \tab 147 \tab 443 \tab 159 \tab 346 \tab 49 \tab 19 \tab 12 \tab 2348 \cr
}
}

\keyword{variable}
