% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_594.R
\name{relsp2}
\alias{relsp2}
\title{Relationship of person 2 to spouse of househol}
\source{
General Social Survey https://gss.norc.org
}
\description{
relsp2
}
\details{
Question What is (NAME)’s relationship to spouse?
B. Relationship of second person to spouse?
}
\section{Values}{

\itemize{
\item \code{1} hh spouse, partner
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, not specified
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{41} natural/biological child
\item \code{42} adopted child
\item \code{43} stepchild
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5454/vshow}{GSS Data Explorer page for \code{relsp2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab hh spouse, partner \tab no answer \tab not available in this release \tab Total \cr
   2000 \tab 1298 \tab 1519 \tab - \tab - \tab 2817 \cr
   2002 \tab 1387 \tab 1378 \tab - \tab - \tab 2765 \cr
   2004 \tab 1126 \tab 1686 \tab - \tab - \tab 2812 \cr
   2006 \tab 1924 \tab 2586 \tab - \tab - \tab 4510 \cr
   2008 \tab 870 \tab 1152 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 948 \tab 1094 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 864 \tab 1110 \tab - \tab - \tab 1974 \cr
   2014 \tab 1121 \tab 1417 \tab - \tab - \tab 2538 \cr
   2016 \tab 1344 \tab 1523 \tab - \tab - \tab 2867 \cr
   2018 \tab 1069 \tab 1278 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 11951 \tab 14743 \tab 4 \tab 4032 \tab 30730 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}}
}
\concept{Household Members and Structure}
\keyword{variable}
