% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_594.R
\name{relsp3}
\alias{relsp3}
\title{Relationship of person 3 to spouse of househol}
\source{
General Social Survey https://gss.norc.org
}
\description{
relsp3
}
\details{
Question What is (NAME)'s relationship to spouse?
C. Relationship of third person to spouse?
}
\section{Values}{

\itemize{
\item \verb{[1]} hh spouse, partner
\item \verb{[2]} spouse
\item \verb{[3]} partner, fiance(e), boyfriend, girlfriend, etc.
\item \verb{[4]} child, not specified
\item \verb{[5]} grandchild, great-grandchild
\item \verb{[6]} son-in-law, daughter-in-law
\item \verb{[7]} parent
\item \verb{[8]} parent-in-law
\item \verb{[9]} grandparent, great-grandparent
\item \verb{[10]} uncle or aunt
\item \verb{[11]} niece or nephew
\item \verb{[12]} cousin
\item \verb{[13]} brother, sister (full, half, or step sibling)
\item \verb{[14]} sibling-in-law
\item \verb{[15]} other relative (great-aunt, grandniece, etc.)
\item \verb{[16]} roommate, housemate
\item \verb{[17]} friend
\item \verb{[18]} child of non-relative
\item \verb{[19]} other non-relative (employee, boarder)
\item \verb{[41]} natural/biological child
\item \verb{[42]} adopted child
\item \verb{[43]} stepchild
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5455/vshow}{GSS Data Explorer page for \code{relsp3}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab child, not specified \tab grandchild, great-grandchild \tab son-in-law, daughter-in-law \tab parent \tab parent-in-law \tab niece or nephew \tab brother, sister (full, half, or step sibling) \tab sibling-in-law \tab other relative (great-aunt, grandniece, etc.) \tab roommate, housemate \tab friend \tab child of non-relative \tab other non-relative (employee, boarder) \tab natural/biological child \tab adopted child \tab stepchild \tab no answer \tab grandparent, great-grandparent \tab uncle or aunt \tab cousin \tab don't know \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab not available in this release \tab Total \cr
   2000 \tab 1935 \tab 97 \tab 14 \tab 4 \tab 16 \tab 9 \tab 2 \tab 5 \tab 3 \tab 2 \tab 3 \tab 4 \tab 8 \tab 6 \tab 556 \tab 16 \tab 33 \tab 104 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2076 \tab 147 \tab 11 \tab 5 \tab 14 \tab 1 \tab 3 \tab 8 \tab 7 \tab 3 \tab 3 \tab 5 \tab 8 \tab 8 \tab 423 \tab 6 \tab 27 \tab 7 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1976 \tab 270 \tab 13 \tab 6 \tab 8 \tab 7 \tab 2 \tab 2 \tab 4 \tab 1 \tab 4 \tab 4 \tab 7 \tab 8 \tab 463 \tab 9 \tab 23 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2812 \cr
   2006 \tab 3125 \tab 426 \tab 20 \tab 3 \tab 23 \tab 16 \tab 4 \tab 10 \tab 8 \tab 2 \tab 11 \tab 12 \tab 10 \tab 22 \tab 743 \tab 11 \tab 58 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 1400 \tab 265 \tab 12 \tab - \tab 8 \tab 5 \tab 4 \tab 8 \tab 2 \tab - \tab 8 \tab 4 \tab 10 \tab 4 \tab 261 \tab 3 \tab 25 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 1508 \tab 270 \tab 6 \tab 3 \tab 6 \tab 2 \tab 2 \tab 9 \tab 4 \tab 1 \tab 11 \tab 5 \tab 9 \tab 4 \tab 176 \tab 6 \tab 16 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1418 \tab 262 \tab 6 \tab 1 \tab 3 \tab 8 \tab 4 \tab 11 \tab 3 \tab 2 \tab 9 \tab 4 \tab 7 \tab 4 \tab 200 \tab 3 \tab 22 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 1974 \cr
   2014 \tab 1854 \tab 350 \tab 7 \tab 3 \tab 10 \tab 10 \tab 4 \tab 4 \tab 5 \tab - \tab 14 \tab 5 \tab 7 \tab 4 \tab 226 \tab 4 \tab 21 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab 2143 \tab 462 \tab 15 \tab 3 \tab 14 \tab 10 \tab 5 \tab 10 \tab 3 \tab 3 \tab 9 \tab 7 \tab 13 \tab 9 \tab 116 \tab 7 \tab 35 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 1759 \tab 327 \tab 6 \tab 2 \tab 3 \tab 13 \tab 2 \tab 4 \tab 7 \tab - \tab 5 \tab 1 \tab 7 \tab 3 \tab 178 \tab 4 \tab 21 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 19194 \tab 2876 \tab 110 \tab 30 \tab 105 \tab 81 \tab 32 \tab 71 \tab 46 \tab 14 \tab 77 \tab 51 \tab 86 \tab 72 \tab 3342 \tab 69 \tab 281 \tab 120 \tab 5 \tab 11 \tab 9 \tab 15 \tab 1 \tab 4032 \tab 30730 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp2}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}}
}
\concept{Household Members and Structure}
\keyword{variable}
