% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{relsprt}
\alias{relsprt}
\title{Is r spiritual person}
\source{
General Social Survey https://gss.norc.org
}
\description{
relsprt
}
\details{
Question What best describes you:
}
\section{Values}{

\itemize{
\item \verb{[1]} i follow a religion and consider myself to be a spiritual person interested in the sacred and the supernatural
\item \verb{[2]} i follow a religion, but don't consider myself to be a spiritual person interested in the sacred and the supernat
\item \verb{[3]} i don't follow a religion, but consider myself to be a spiritual person interested in the sacred and the supernat
\item \verb{[4]} i don't follow a religion and don't consider myself to be a spiritual person interested in the sacred and the sup
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4711/vshow}{GSS Data Explorer page for \code{relsprt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab i follow a religion and consider myself to be a spiritual person interested in the sacred and the supernatural \tab i follow a religion, but don't consider myself to be a spiritual person interested in the sacred and the supernat \tab i don't follow a religion, but consider myself to be a spiritual person interested in the sacred and the supernat \tab i don't follow a religion and don't consider myself to be a spiritual person interested in the sacred and the sup \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 528 \tab 304 \tab 312 \tab 154 \tab 48 \tab 19 \tab 2023 \cr
   2018 \tab 1173 \tab 495 \tab 231 \tab 243 \tab 145 \tab 44 \tab 17 \tab 2348 \cr
   Total \tab 1831 \tab 1023 \tab 535 \tab 555 \tab 299 \tab 92 \tab 36 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{rightpeo}},
\code{\link{vistholy}}
}
\concept{ISSP Religion}
\keyword{variable}
