% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{reltrad}
\alias{reltrad}
\title{Religious tradition}
\source{
General Social Survey https://gss.norc.org
}
\description{
reltrad
}
\details{
Question RELTRAD is constructed from variables RELIG, DENOM, OTHER, and ATTEND. The missing values coded as ".z: Variable-specific reserve code" represent OTHER denominations that are not originally covered by the Steensland et al. (2000) article (i.e., OTHER values larger than 148). Please see Appendix F in the codebook for more information.
}
\section{Values}{

\itemize{
\item \verb{[1]} evangelical protestant
\item \verb{[2]} mainline protestant
\item \verb{[3]} black protestant
\item \verb{[4]} catholic
\item \verb{[5]} jewish
\item \verb{[6]} other faith
\item \verb{[7]} nonaffiliated
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8253/vshow}{GSS Data Explorer page for \code{reltrad}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab evangelical protestant \tab mainline protestant \tab black protestant \tab catholic \tab jewish \tab other faith \tab nonaffiliated \tab iap \tab no answer \tab don't know \tab see codebook \tab skipped on web \tab Total \cr
   1972 \tab 291 \tab 467 \tab 213 \tab 413 \tab 54 \tab 60 \tab 83 \tab 25 \tab 7 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 314 \tab 427 \tab 136 \tab 388 \tab 42 \tab 56 \tab 96 \tab 34 \tab 6 \tab 5 \tab - \tab - \tab 1504 \cr
   1974 \tab 331 \tab 431 \tab 133 \tab 376 \tab 44 \tab 31 \tab 101 \tab 27 \tab 2 \tab 8 \tab - \tab - \tab 1484 \cr
   1975 \tab 319 \tab 461 \tab 126 \tab 363 \tab 23 \tab 41 \tab 113 \tab 38 \tab 3 \tab 3 \tab - \tab - \tab 1490 \cr
   1976 \tab 345 \tab 421 \tab 103 \tab 390 \tab 27 \tab 49 \tab 114 \tab 48 \tab 2 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 349 \tab 444 \tab 138 \tab 373 \tab 35 \tab 48 \tab 93 \tab 43 \tab 7 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 364 \tab 423 \tab 131 \tab 384 \tab 29 \tab 51 \tab 119 \tab 26 \tab 5 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 357 \tab 399 \tab 105 \tab 362 \tab 32 \tab 72 \tab 105 \tab 32 \tab 4 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 358 \tab 450 \tab 366 \tab 399 \tab 38 \tab 56 \tab 132 \tab 48 \tab 13 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 357 \tab 412 \tab 121 \tab 439 \tab 43 \tab 83 \tab 117 \tab 21 \tab 6 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 394 \tab 322 \tab 123 \tab 375 \tab 27 \tab 74 \tab 107 \tab 36 \tab 14 \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 434 \tab 334 \tab 110 \tab 408 \tab 32 \tab 71 \tab 109 \tab 29 \tab 7 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 401 \tab 301 \tab 130 \tab 380 \tab 38 \tab 85 \tab 98 \tab 32 \tab 4 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 433 \tab 341 \tab 379 \tab 376 \tab 21 \tab 103 \tab 121 \tab 32 \tab 13 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 404 \tab 289 \tab 125 \tab 384 \tab 30 \tab 91 \tab 118 \tab 32 \tab 7 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 409 \tab 334 \tab 116 \tab 386 \tab 23 \tab 82 \tab 120 \tab 59 \tab 8 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 348 \tab 322 \tab 109 \tab 327 \tab 27 \tab 79 \tab 109 \tab 41 \tab 8 \tab 1 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 427 \tab 310 \tab 143 \tab 386 \tab 32 \tab 83 \tab 102 \tab 30 \tab 4 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 478 \tab 321 \tab 121 \tab 351 \tab 33 \tab 87 \tab 146 \tab 58 \tab 10 \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab 787 \tab 589 \tab 237 \tab 759 \tab 59 \tab 200 \tab 274 \tab 64 \tab 21 \tab 1 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 732 \tab 549 \tab 261 \tab 685 \tab 68 \tab 199 \tab 339 \tab 55 \tab 14 \tab 1 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 710 \tab 483 \tab 230 \tab 705 \tab 50 \tab 128 \tab 396 \tab 75 \tab 48 \tab 6 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 673 \tab 470 \tab 262 \tab 679 \tab 63 \tab 153 \tab 398 \tab 97 \tab 18 \tab 2 \tab 2 \tab - \tab 2817 \cr
   2002 \tab 644 \tab 463 \tab 236 \tab 673 \tab 48 \tab 159 \tab 379 \tab 113 \tab 22 \tab 26 \tab 2 \tab - \tab 2765 \cr
   2004 \tab 745 \tab 410 \tab 208 \tab 656 \tab 55 \tab 179 \tab 403 \tab 125 \tab 18 \tab 5 \tab 8 \tab - \tab 2812 \cr
   2006 \tab 1088 \tab 627 \tab 366 \tab 1115 \tab 78 \tab 229 \tab 739 \tab 195 \tab 33 \tab 18 \tab 22 \tab - \tab 4510 \cr
   2008 \tab 480 \tab 283 \tab 153 \tab 470 \tab 39 \tab 128 \tab 332 \tab 104 \tab 11 \tab 18 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 484 \tab 275 \tab 166 \tab 482 \tab 37 \tab 103 \tab 363 \tab 103 \tab 16 \tab 9 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 501 \tab 250 \tab 136 \tab 444 \tab 28 \tab 109 \tab 387 \tab 100 \tab 4 \tab 9 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 582 \tab 315 \tab 170 \tab 606 \tab 40 \tab 153 \tab 522 \tab 120 \tab 15 \tab 10 \tab 5 \tab - \tab 2538 \cr
   2016 \tab 679 \tab 303 \tab 208 \tab 649 \tab 51 \tab 156 \tab 620 \tab 162 \tab 22 \tab 12 \tab 5 \tab - \tab 2867 \cr
   2018 \tab 534 \tab 277 \tab 162 \tab 493 \tab 39 \tab 132 \tab 542 \tab 120 \tab 20 \tab 20 \tab 9 \tab - \tab 2348 \cr
   2021 \tab 715 \tab 470 \tab 169 \tab 824 \tab 76 \tab 270 \tab 1121 \tab 279 \tab 46 \tab 20 \tab 3 \tab 39 \tab 4032 \cr
   2022 \tab 609 \tab 354 \tab 196 \tab 744 \tab 61 \tab 252 \tab 977 \tab 284 \tab 32 \tab 20 \tab 6 \tab 9 \tab 3544 \cr
   2024 \tab 622 \tab 308 \tab 212 \tab 649 \tab 46 \tab 276 \tab 876 \tab 270 \tab 6 \tab 26 \tab 6 \tab 12 \tab 3309 \cr
   Total \tab 17698 \tab 13635 \tab 6300 \tab 17893 \tab 1468 \tab 4128 \tab 10771 \tab 2957 \tab 476 \tab 224 \tab 89 \tab 60 \tab 75699 \cr
}
}

\seealso{
Other Religion: 
\code{\link{conorelgsp}}
}
\concept{Religion}
\keyword{variable}
