% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{repairs}
\alias{repairs}
\title{Who in household does small repair's}
\source{
General Social Survey https://gss.norc.org
}
\description{
repairs
}
\details{
Question In your household who does the following things . . .
B. Makes small repairs around the house.
}
\section{Values}{

\itemize{
\item \code{1} always the woman
\item \code{2} usually the woman
\item \code{3} about equal or both together
\item \code{4} usually the man
\item \code{5} always the man
\item \code{6} is done by a third person
\item \code{7} not married
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4419/vshow}{GSS Data Explorer page for \code{repairs}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
