% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{repairs1}
\alias{repairs1}
\title{Who in household does small repairs}
\source{
General Social Survey https://gss.norc.org
}
\description{
repairs1
}
\details{
Question In your household who does the following things...?
B. Makes small repairs around the house?
}
\section{Values}{

\itemize{
\item \verb{[1]} always me
\item \verb{[2]} usually me
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually my spouse/partner
\item \verb{[5]} always my spouse/partner
\item \verb{[6]} is done by a third person
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4430/vshow}{GSS Data Explorer page for \code{repairs1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab always me \tab usually me \tab about equal or both together \tab usually my spouse/partner \tab always my spouse/partner \tab is done by a third person \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2058 \tab 142 \tab 147 \tab 129 \tab 219 \tab 15 \tab 30 \tab 3 \tab 22 \tab 2765 \cr
   2012 \tab 1274 \tab 185 \tab 117 \tab 91 \tab 156 \tab 111 \tab 33 \tab 5 \tab 2 \tab 1974 \cr
   Total \tab 3332 \tab 327 \tab 264 \tab 220 \tab 375 \tab 126 \tab 63 \tab 8 \tab 24 \tab 4739 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{shopfood}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
