% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{replaceu}
\alias{replaceu}
\title{How easy r's firm replaces r if r left}
\source{
General Social Survey https://gss.norc.org
}
\description{
replaceu
}
\details{
Question B. How difficult or easy do you think it would be for your firm or organization to replace you if you left?
}
\section{Values}{

\itemize{
\item \verb{[1]} very easy
\item \verb{[2]} fairly easy
\item \verb{[3]} neither easy nor difficult
\item \verb{[4]} fairly difficult
\item \verb{[5]} very difficult
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4619/vshow}{GSS Data Explorer page for \code{replaceu}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very easy \tab fairly easy \tab neither easy nor difficult \tab fairly difficult \tab very difficult \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 205 \tab 256 \tab 157 \tab 256 \tab 131 \tab 10 \tab 3 \tab 4510 \cr
   Total \tab 3492 \tab 205 \tab 256 \tab 157 \tab 256 \tab 131 \tab 10 \tab 3 \tab 4510 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
