% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_318.R
\name{reptile}
\alias{reptile}
\title{Reptile (turtle, snake, lizard)}
\source{
General Social Survey https://gss.norc.org
}
\description{
reptile
}
\details{
Question Current pet - Reptile
}
\section{Values}{

\itemize{
\item \verb{[0]} does not have
\item \verb{[1]} has
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7202/vshow}{GSS Data Explorer page for \code{reptile}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab does not have \tab has \tab no answer \tab Total \cr
   2018 \tab 1649 \tab 643 \tab 30 \tab 26 \tab 2348 \cr
   Total \tab 1649 \tab 643 \tab 30 \tab 26 \tab 2348 \cr
}
}

\keyword{variable}
