% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{res16}
\alias{res16}
\title{Type of place lived in when 16 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
res16
}
\details{
Question Which of the categories on this card comes closest to the type of place you were living in when you were 16 years old?
}
\section{Values}{

\itemize{
\item \code{1} in open country but not on a farm
\item \code{2} farm
\item \code{3} in a small city or town (under 50,000)
\item \code{4} in a medium-size city (50,000-250,000)
\item \code{5} in a suburb near a large city
\item \code{6} in a large city (over 250,000)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/83/vshow}{GSS Data Explorer page for \code{res16}}.
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\concept{Residence}
\keyword{variable}
