% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{res161}
\alias{res161}
\title{Where spend childhood until age 16?}
\source{
General Social Survey https://gss.norc.org
}
\description{
res161
}
\details{
Question Where did you spend most of your childhood until age 16?
}
\section{Values}{

\itemize{
\item \code{1} in this community
\item \code{2} in a different community same state
\item \code{3} in a different state in america
\item \code{4} outside america
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4865/vshow}{GSS Data Explorer page for \code{res161}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab in this community \tab in a different community same state \tab in a different state in america \tab outside america \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 418 \tab 427 \tab 416 \tab 81 \tab 1 \tab 24 \tab 2904 \cr
   Total \tab 1537 \tab 418 \tab 427 \tab 416 \tab 81 \tab 1 \tab 24 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{refugees}}
}
\concept{ISSP National Identity}
\keyword{variable}
