% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{res2006}
\alias{res2006}
\title{Was r living in usa during march-may 2006}
\source{
General Social Survey https://gss.norc.org
}
\description{
res2006
}
\details{
Question A. During March through May 2006, were you living in the United States or were you living outside of the country?
}
\section{Values}{

\itemize{
\item \code{1} living in the u.s.
\item \code{2} not living in u.s.
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5585/vshow}{GSS Data Explorer page for \code{res2006}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
