% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{res2008}
\alias{res2008}
\title{Was r living in usa during april-june 2008}
\source{
General Social Survey https://gss.norc.org
}
\description{
res2008
}
\details{
Question B. During April through June 2008, were you living in the United States or were you living outside of the country?
}
\section{Values}{

\itemize{
\item \verb{[1]} living in the u.s.
\item \verb{[2]} not living in u.s.
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5586/vshow}{GSS Data Explorer page for \code{res2008}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab living in the u.s. \tab not living in u.s. \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2010 \tab - \tab 2023 \tab 20 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 2 \tab 1931 \tab 40 \tab - \tab 1 \tab - \tab 1974 \cr
   Total \tab 2 \tab 3954 \tab 60 \tab 1 \tab 1 \tab 0 \tab 4018 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{batch}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{cshutyp12}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{subsamprate}},
\code{\link{version}},
\code{\link{vpsu}},
\code{\link{vstrat}},
\code{\link{wtss}},
\code{\link{wtssall}},
\code{\link{wtssnr}}

Other Dwelling (R's): 
\code{\link{bornhome}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{dwelcity}},
\code{\link{dwell5}},
\code{\link{dwelling}},
\code{\link{dwelling_8222}},
\code{\link{dwelngh}},
\code{\link{dwelown}},
\code{\link{evdwell}},
\code{\link{geomobil}},
\code{\link{huclean}},
\code{\link{res2006}},
\code{\link{res2010}}
}
\concept{Dwelling (R's)}
\concept{Sampling and Weights}
\keyword{variable}
