% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{res2010}
\alias{res2010}
\title{Was r living in usa during april-june 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
res2010
}
\details{
Question C. During April through June 2010, were you living in the United States or were you living outside of the country?
}
\section{Values}{

\itemize{
\item \verb{[1]} living in the u.s.
\item \verb{[2]} not living in u.s.
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5587/vshow}{GSS Data Explorer page for \code{res2010}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab living in the u.s. \tab not living in u.s. \tab no answer \tab Total \cr
   2012 \tab 2 \tab 1951 \tab 20 \tab 1 \tab 1974 \cr
   2014 \tab - \tab 2483 \tab 54 \tab 1 \tab 2538 \cr
   Total \tab 2 \tab 4434 \tab 74 \tab 2 \tab 4512 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2012}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
