% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_92.R
\name{resdefwk}
\alias{resdefwk}
\title{Local economy depends on defense work?}
\source{
General Social Survey https://gss.norc.org
}
\description{
resdefwk
}
\details{
Question Would you say the economy of (NAME METROPOLITAN AREA OR COUNTY) is very dependent on defense business, somewhat dependent, or not dependent at all on defense business?
}
\section{Values}{

\itemize{
\item \code{1} very dependent
\item \code{2} somewhat dependent
\item \code{3} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/833/vshow}{GSS Data Explorer page for \code{resdefwk}}.
}

\seealso{
Other Core: 
\code{\link{defwrknw}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obvol}},
\code{\link{obvote}}

Other Military: 
\code{\link{defwrknw}},
\code{\link{milwrkev}},
\code{\link{milwrknw}}
}
\concept{Core}
\concept{Military}
\keyword{variable}
