% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_520.R
\name{resnatur}
\alias{resnatur}
\title{Humans should respect naturecreated by god}
\source{
General Social Survey https://gss.norc.org
}
\description{
resnatur
}
\details{
Question And please check one box for each of these statements to show how much you agree or disagree with it.
D. Human beings should respect nature because it was created by God.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4728/vshow}{GSS Data Explorer page for \code{resnatur}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 387 \tab 747 \tab 238 \tab 63 \tab 46 \tab 23 \tab 53 \tab 1606 \cr
   1994 \tab 1606 \tab 362 \tab 663 \tab 196 \tab 64 \tab 35 \tab 24 \tab 42 \tab 2992 \cr
   Total \tab 1655 \tab 749 \tab 1410 \tab 434 \tab 127 \tab 81 \tab 47 \tab 95 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grntaxes}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}}
}
\concept{ISSP Environment}
\keyword{variable}
