% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{restact}
\alias{restact}
\title{Afterlife contrast: peaceful or action}
\source{
General Social Survey https://gss.norc.org
}
\description{
restact
}
\details{
Question A. People picture life after death in many different ways. We'd like to know how you think of life after death. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of life after death. Look at the first set of contrasting images. If you think that life after death is "A life like the one here on earth only better," you would place yourself at 1. If you think it is "A life without many things which make our present life enjoyable" you would place yourself at 7. If you think it is somewhere between these two you would place yourself at 2, 3, 4, 5, or 6. Where would you place your images of life after death for . . .?
}
\section{Values}{

\itemize{
\item \verb{[1]} peaceful
\item \verb{[7]} intense action
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/313/vshow}{GSS Data Explorer page for \code{restact}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab peaceful \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab intense action \tab don't know \tab no answer \tab Total \cr
   1986 \tab 246 \tab 643 \tab 167 \tab 109 \tab 182 \tab 19 \tab 15 \tab 48 \tab 17 \tab 24 \tab 1470 \cr
   Total \tab 246 \tab 643 \tab 167 \tab 109 \tab 182 \tab 19 \tab 15 \tab 48 \tab 17 \tab 24 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
