% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_183.R
\name{restless}
\alias{restless}
\title{Felt restless?}
\source{
General Social Survey https://gss.norc.org
}
\description{
restless
}
\details{
Question Now I'm going to read a list of different feelings that people sometimes have. After each one, I would like you to tell me on how many days you have felt this way during the past 7 days. On how many days in the past 7 days have you... (SPECIFY NUMBER OF DAYS)
(IF RESPONDENT ASKS, SAY THAT FEELING THIS WAY AT ANY TIME DURING A DAY MAKES THAT DAY "COUNT.")
N. Felt so restless that you couldn't sit long in a chair?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1719/vshow}{GSS Data Explorer page for \code{restless}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 811 \tab 180 \tab 98 \tab 90 \tab 50 \tab 69 \tab 23 \tab 123 \tab 6 \tab 10 \tab 2904 \cr
   Total \tab 1444 \tab 811 \tab 180 \tab 98 \tab 90 \tab 50 \tab 69 \tab 23 \tab 123 \tab 6 \tab 10 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{anxious}},
\code{\link{atease}},
\code{\link{contentd}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{ovrjoyed}},
\code{\link{worried}}
}
\concept{Emotions}
\keyword{variable}
