% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_323.R
\name{retchnge}
\alias{retchnge}
\title{R returned money after getting too much change}
\source{
General Social Survey https://gss.norc.org
}
\description{
retchnge
}
\details{
Question During the past 12 months, how often have you done each of the following things:
C. Returned money to a cashier after getting too much change
}
\section{Values}{

\itemize{
\item \verb{[1]} more than once a week
\item \verb{[2]} once a week
\item \verb{[3]} once a month
\item \verb{[4]} at least 2 or 3 times in the past year
\item \verb{[5]} once in the past year
\item \verb{[6]} not at all in the past year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2879/vshow}{GSS Data Explorer page for \code{retchnge}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab more than once a week \tab once a week \tab once a month \tab at least 2 or 3 times in the past year \tab once in the past year \tab not at all in the past year \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 8 \tab 6 \tab 28 \tab 256 \tab 330 \tab 709 \tab 30 \tab 5 \tab 2765 \cr
   2004 \tab 1472 \tab 14 \tab 14 \tab 42 \tab 266 \tab 359 \tab 637 \tab 5 \tab 3 \tab 2812 \cr
   2012 \tab 666 \tab 13 \tab 7 \tab 23 \tab 155 \tab 324 \tab 772 \tab 8 \tab 6 \tab 1974 \cr
   2014 \tab 1264 \tab 11 \tab 14 \tab 22 \tab 159 \tab 249 \tab 809 \tab 5 \tab 5 \tab 2538 \cr
   Total \tab 4795 \tab 46 \tab 41 \tab 115 \tab 836 \tab 1262 \tab 2927 \tab 48 \tab 19 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{cutahead}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{givblood}},
\code{\link{givchrty}},
\code{\link{givhmlss}},
\code{\link{givseat}},
\code{\link{volchrty}}
}
\concept{Altruism}
\keyword{variable}
