% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{revpub}
\alias{revpub}
\title{Allow revolutionaries to publish books}
\source{
General Social Survey https://gss.norc.org
}
\description{
revpub
}
\details{
Question There are some people whose views are considered extreme by the majority. First, consider people who want to overthrow the government by revolution. please circle a number after each question to show if you think such people should be allowed to . . .
C. Publish books expressing their views.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely allowed
\item \verb{[2]} probably allowed
\item \verb{[3]} probably not allowed
\item \verb{[4]} definitely not allowed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3990/vshow}{GSS Data Explorer page for \code{revpub}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 170 \tab 182 \tab 88 \tab 186 \tab 37 \tab 14 \tab 1534 \cr
   1990 \tab 155 \tab 365 \tab 388 \tab 167 \tab 215 \tab 58 \tab 24 \tab 1372 \cr
   1996 \tab 1572 \tab 407 \tab 443 \tab 156 \tab 218 \tab 67 \tab 41 \tab 2904 \cr
   2006 \tab 2992 \tab 839 \tab 439 \tab 125 \tab 102 \tab 13 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 719 \tab 418 \tab 132 \tab 92 \tab 18 \tab 11 \tab 2867 \cr
   Total \tab 7053 \tab 2500 \tab 1870 \tab 668 \tab 813 \tab 193 \tab 90 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revspeak}},
\code{\link{revtch15}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
