% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{revspeak}
\alias{revspeak}
\title{Allow revolutionary public meetings}
\source{
General Social Survey https://gss.norc.org
}
\description{
revspeak
}
\details{
Question There are some people whose views are considered extreme by the majority. First, consider people who want to overthrow the government by revolution. please circle a number after each question to show if you think such people should be allowed to . . .
A. Hold public meetings to express their views.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3988/vshow}{GSS Data Explorer page for \code{revspeak}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revtch15}}

Other Civil Liberties: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revtch15}}

Other Demonstrations: 
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revtch15}}
}
\concept{Civil Liberties}
\concept{Demonstrations}
\concept{ISSP Role of Gov't}
\keyword{variable}
