% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_499.R
\name{rexpernc}
\alias{rexpernc}
\title{Importance of experience}
\source{
General Social Survey https://gss.norc.org
}
\description{
rexpernc
}
\details{
Question Think of two people doing the same kind of work in your place. What Do you personally think is important in deciding
Looking at the things below, please write ''1'' in the box next to the thing that you think is most important at your place of work. Then write ''2'' next to the thing that you think is next most important, and ''3'' next to the thing that you think is third most important. Leave the other boxes blank.
At my work, deciding on two people doing the same kind of work, how important is...
D. The experience of the employer doing the work
}
\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} 2nd most impt
\item \code{3} 3rd most impt
\item \code{4} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4536/vshow}{GSS Data Explorer page for \code{rexpernc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most impt \tab 2nd most impt \tab 3rd most impt \tab not chosen \tab don't know \tab no answer \tab Total \cr
   1989 \tab 639 \tab 62 \tab 323 \tab 164 \tab 150 \tab 25 \tab 174 \tab 1537 \cr
   Total \tab 639 \tab 62 \tab 323 \tab 164 \tab 150 \tab 25 \tab 174 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{hrsmoney}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rsecjob}},
\code{\link{ryrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
