% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{rfamlook}
\alias{rfamlook}
\title{Hours r spends looking after family members}
\source{
General Social Survey https://gss.norc.org
}
\description{
rfamlook
}
\details{
Question A. On average, how many hours a week do you spend looking after family members (e.g. children, elderly, ill or disabled family members)?
}
\section{Values}{

\itemize{
\item \verb{[97]} 97 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4483/vshow}{GSS Data Explorer page for \code{rfamlook}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 14 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 21 \tab 24 \tab 25 \tab 26 \tab 28 \tab 3 \tab 30 \tab 31 \tab 32 \tab 35 \tab 36 \tab 38 \tab 4 \tab 40 \tab 42 \tab 45 \tab 48 \tab 49 \tab 5 \tab 50 \tab 53 \tab 55 \tab 56 \tab 6 \tab 60 \tab 63 \tab 64 \tab 68 \tab 69 \tab 7 \tab 70 \tab 72 \tab 73 \tab 74 \tab 75 \tab 78 \tab 8 \tab 80 \tab 84 \tab 9 \tab 90 \tab 91 \tab 95 \tab 96 \tab 97 or more \tab no answer \tab 11 \tab 13 \tab 17 \tab 22 \tab 33 \tab 39 \tab 65 \tab 93 \tab 94 \tab don't know \tab skipped on web \tab not available in this year \tab Total \cr
   2012 \tab 672 \tab 543 \tab 28 \tab 49 \tab 21 \tab 11 \tab 29 \tab 7 \tab 2 \tab 44 \tab 46 \tab 2 \tab 23 \tab 18 \tab 3 \tab 5 \tab 27 \tab 24 \tab 1 \tab 3 \tab 9 \tab 3 \tab 1 \tab 37 \tab 50 \tab 5 \tab 3 \tab 6 \tab 3 \tab 34 \tab 19 \tab 1 \tab 2 \tab 2 \tab 21 \tab 25 \tab 2 \tab 1 \tab 2 \tab 1 \tab 10 \tab 13 \tab 6 \tab 2 \tab 1 \tab 4 \tab 1 \tab 39 \tab 14 \tab 4 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 81 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2022 \tab 2385 \tab 518 \tab 28 \tab 31 \tab 15 \tab 12 \tab 15 \tab 1 \tab 2 \tab 67 \tab 42 \tab 4 \tab 23 \tab 13 \tab 1 \tab 2 \tab 28 \tab 17 \tab - \tab - \tab 8 \tab 3 \tab 1 \tab 26 \tab 28 \tab - \tab - \tab 4 \tab - \tab 39 \tab 9 \tab - \tab 1 \tab 7 \tab 27 \tab 15 \tab - \tab - \tab 1 \tab - \tab 12 \tab 4 \tab 4 \tab - \tab - \tab - \tab - \tab 31 \tab 11 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 40 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 24 \tab 23 \tab - \tab 3544 \cr
   Total \tab 3057 \tab 1061 \tab 56 \tab 80 \tab 36 \tab 23 \tab 44 \tab 8 \tab 4 \tab 111 \tab 88 \tab 6 \tab 46 \tab 31 \tab 4 \tab 7 \tab 55 \tab 41 \tab 1 \tab 3 \tab 17 \tab 6 \tab 2 \tab 63 \tab 78 \tab 5 \tab 3 \tab 10 \tab 3 \tab 73 \tab 28 \tab 1 \tab 3 \tab 9 \tab 48 \tab 40 \tab 2 \tab 1 \tab 3 \tab 1 \tab 22 \tab 17 \tab 10 \tab 2 \tab 1 \tab 4 \tab 1 \tab 70 \tab 25 \tab 6 \tab 8 \tab 3 \tab 1 \tab 2 \tab 4 \tab 121 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 24 \tab 23 \tab 0 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{buythngs}},
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{chldcare}},
\code{\link{clean1}},
\code{\link{cohabfst}},
\code{\link{cohabit}},
\code{\link{cohabok}},
\code{\link{cooking1}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{deckids}},
\code{\link{dinner}},
\code{\link{divbest}},
\code{\link{divhubby}},
\code{\link{divifkd1}},
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokd1}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{eldcost}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{evdiv}},
\code{\link{fambudgt}},
\code{\link{famsuffr}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{fejobind}},
\code{\link{femarry}},
\code{\link{fewknokd}},
\code{\link{fewrksup}},
\code{\link{fourkids}},
\code{\link{hapifwrk}},
\code{\link{happy7}},
\code{\link{hhclean1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}},
\code{\link{kidsocst}},
\code{\link{kidsuffr}},
\code{\link{kidtrble}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{mardiv}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{mawork14}},
\code{\link{mawrkwrm}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{memarry}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{rhhwork}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{sexhar}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{singlpar}},
\code{\link{spborn}},
\code{\link{spevdiv}},
\code{\link{spfalook}},
\code{\link{sphhwork}},
\code{\link{splive}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{teensex1}},
\code{\link{threkids}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhm1}},
\code{\link{tiredhme}},
\code{\link{tiredwk1}},
\code{\link{tiredwrk}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{twokids}},
\code{\link{weekend}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}

Other Family: 
\code{\link{adoptkid}},
\code{\link{agewed}},
\code{\link{agewedcr}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{balwkfam}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{chldben}},
\code{\link{chlddth}},
\code{\link{chldhrt}},
\code{\link{clsclsex}},
\code{\link{cousins}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce}},
\code{\link{engda}},
\code{\link{engson}},
\code{\link{evdiv}},
\code{\link{evkid}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{famlife}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famvswk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{feovrwrk}},
\code{\link{flextime}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{inlawnum}},
\code{\link{inperson}},
\code{\link{jobvsfam}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidcall}},
\code{\link{kidlived}},
\code{\link{kidmost}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{kidvisit}},
\code{\link{letters}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{madeath}},
\code{\link{malive}},
\code{\link{malive1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{marital}},
\code{\link{marunion}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{meetings}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{mykids}},
\code{\link{niecenep}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{numcntct}},
\code{\link{numprobs}},
\code{\link{obtohelp}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{othkids}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{padeath}},
\code{\link{palive}},
\code{\link{palive1}},
\code{\link{parleave}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{pawrkgrw}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{posslq}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relcall}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relmarry}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibdeath}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibvisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{spdeath}},
\code{\link{spfalook}},
\code{\link{spfam16}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{sppart}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{widowed}},
\code{\link{wkvsfam}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}}
}
\concept{Family}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
