% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{rheight}
\alias{rheight}
\title{Respondent height}
\source{
General Social Survey https://gss.norc.org
}
\description{
rheight
}
\details{
Question What is your height? Reported in inches.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7971/vshow}{GSS Data Explorer page for \code{rheight}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab 53 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 80 \tab 95 \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2382 \tab 1 \tab 1 \tab 1 \tab 2 \tab 17 \tab 36 \tab 42 \tab 87 \tab 93 \tab 86 \tab 91 \tab 107 \tab 90 \tab 78 \tab 70 \tab 83 \tab 76 \tab 71 \tab 28 \tab 28 \tab 14 \tab 14 \tab 7 \tab 2 \tab 1 \tab 1 \tab 3 \tab 12 \tab 20 \tab 3544 \cr
   Total \tab 2382 \tab 1 \tab 1 \tab 1 \tab 2 \tab 17 \tab 36 \tab 42 \tab 87 \tab 93 \tab 86 \tab 91 \tab 107 \tab 90 \tab 78 \tab 70 \tab 83 \tab 76 \tab 71 \tab 28 \tab 28 \tab 14 \tab 14 \tab 7 \tab 2 \tab 1 \tab 1 \tab 3 \tab 12 \tab 20 \tab 3544 \cr
}
}

\keyword{variable}
