% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{rhhwork}
\alias{rhhwork}
\title{How many hours a week does r spend on hh work}
\source{
General Social Survey https://gss.norc.org
}
\description{
rhhwork
}
\details{
Question A. On average, how many hours a week do you personally spend on household work, not including childcare  and leisure time activities?
}
\section{Values}{

\itemize{
\item \verb{[97]} 97 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4435/vshow}{GSS Data Explorer page for \code{rhhwork}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 28 \tab 3 \tab 30 \tab 34 \tab 35 \tab 36 \tab 4 \tab 40 \tab 45 \tab 48 \tab 5 \tab 50 \tab 51 \tab 55 \tab 56 \tab 6 \tab 60 \tab 7 \tab 72 \tab 8 \tab 84 \tab 9 \tab don't know \tab no answer \tab 11 \tab 17 \tab 27 \tab 42 \tab 49 \tab 70 \tab 80 \tab 85 \tab 97 or more \tab 26 \tab 32 \tab 47 \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 2058 \tab 11 \tab 17 \tab 110 \tab 26 \tab 1 \tab 11 \tab 27 \tab 8 \tab 1 \tab 1 \tab 54 \tab 36 \tab 5 \tab 1 \tab 1 \tab 4 \tab 10 \tab 6 \tab 59 \tab 17 \tab 1 \tab 9 \tab 1 \tab 55 \tab 11 \tab 2 \tab 2 \tab 71 \tab 4 \tab 1 \tab 1 \tab 1 \tab 39 \tab 3 \tab 30 \tab 1 \tab 42 \tab 1 \tab 2 \tab 2 \tab 22 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2012 \tab 672 \tab 46 \tab 65 \tab 146 \tab 32 \tab 2 \tab 46 \tab 63 \tab 7 \tab 3 \tab 1 \tab 108 \tab 75 \tab 13 \tab 2 \tab - \tab 7 \tab 18 \tab 10 \tab 101 \tab 20 \tab - \tab 11 \tab - \tab 110 \tab 20 \tab - \tab 3 \tab 136 \tab 9 \tab - \tab - \tab 1 \tab 67 \tab 1 \tab 60 \tab 2 \tab 69 \tab - \tab 6 \tab 21 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2022 \tab 2385 \tab 26 \tab 34 \tab 137 \tab 35 \tab 2 \tab 42 \tab 37 \tab 5 \tab 3 \tab - \tab 107 \tab 65 \tab 14 \tab 3 \tab - \tab 11 \tab 12 \tab 9 \tab 103 \tab 19 \tab - \tab 5 \tab - \tab 93 \tab 17 \tab 2 \tab - \tab 123 \tab 2 \tab - \tab - \tab 1 \tab 59 \tab 4 \tab 58 \tab 2 \tab 79 \tab - \tab 6 \tab 13 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 15 \tab - \tab 3544 \cr
   Total \tab 5115 \tab 83 \tab 116 \tab 393 \tab 93 \tab 5 \tab 99 \tab 127 \tab 20 \tab 7 \tab 2 \tab 269 \tab 176 \tab 32 \tab 6 \tab 1 \tab 22 \tab 40 \tab 25 \tab 263 \tab 56 \tab 1 \tab 25 \tab 1 \tab 258 \tab 48 \tab 4 \tab 5 \tab 330 \tab 15 \tab 1 \tab 1 \tab 3 \tab 165 \tab 8 \tab 148 \tab 5 \tab 190 \tab 1 \tab 14 \tab 36 \tab 31 \tab 2 \tab 2 \tab 3 \tab 3 \tab 1 \tab 4 \tab 3 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 15 \tab 0 \tab 8283 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{buythngs}},
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{chldcare}},
\code{\link{clean1}},
\code{\link{cohabfst}},
\code{\link{cohabit}},
\code{\link{cohabok}},
\code{\link{cooking1}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{deckids}},
\code{\link{dinner}},
\code{\link{divbest}},
\code{\link{divhubby}},
\code{\link{divifkd1}},
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokd1}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{eldcost}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{evdiv}},
\code{\link{fambudgt}},
\code{\link{famsuffr}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{fejobind}},
\code{\link{femarry}},
\code{\link{fewknokd}},
\code{\link{fewrksup}},
\code{\link{fourkids}},
\code{\link{hapifwrk}},
\code{\link{happy7}},
\code{\link{hhclean1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}},
\code{\link{kidsocst}},
\code{\link{kidsuffr}},
\code{\link{kidtrble}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{mardiv}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{mawork14}},
\code{\link{mawrkwrm}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{memarry}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{rfamlook}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{sexhar}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{singlpar}},
\code{\link{spborn}},
\code{\link{spevdiv}},
\code{\link{spfalook}},
\code{\link{sphhwork}},
\code{\link{splive}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{teensex1}},
\code{\link{threkids}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhm1}},
\code{\link{tiredhme}},
\code{\link{tiredwk1}},
\code{\link{tiredwrk}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{twokids}},
\code{\link{weekend}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}

Other Household: 
\code{\link{adults}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{dinner}},
\code{\link{earnrs}},
\code{\link{fairhwrk}},
\code{\link{fambudgt}},
\code{\link{famfinan}},
\code{\link{famgen}},
\code{\link{famgen_7522}},
\code{\link{famgen_exp}},
\code{\link{famgen_exp_22}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{fewrksup}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{hefinfo}},
\code{\link{hhclean1}},
\code{\link{hhrace}},
\code{\link{hhtype}},
\code{\link{hhtype1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homeless}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{hubbywk1}},
\code{\link{hwduties}},
\code{\link{incdef}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{mrmom}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{ownhh}},
\code{\link{ownhh1}},
\code{\link{povline}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relhh1}},
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh14}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp11}},
\code{\link{relsp12}},
\code{\link{relsp13}},
\code{\link{relsp14}},
\code{\link{relsp2}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{respnum}},
\code{\link{rhmewrk}},
\code{\link{rplace}},
\code{\link{rplace_7522}},
\code{\link{rvisitor}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{sphhwork}},
\code{\link{sphmewrk}},
\code{\link{strsshme}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{timehome}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{visitors}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{womenben}},
\code{\link{womenhrt}}
}
\concept{Household}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
