% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{rhiinc}
\alias{rhiinc}
\title{R's income is high}
\source{
General Social Survey https://gss.norc.org
}
\description{
rhiinc
}
\details{
Question For each statement about your main job below, please circle one code to show how much you agree or disagree that it applies to your job.
B. My income is high
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4542/vshow}{GSS Data Explorer page for \code{rhiinc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 38 \tab 190 \tab 266 \tab 297 \tab 73 \tab 7 \tab 27 \tab 1537 \cr
   1998 \tab 1967 \tab 40 \tab 162 \tab 248 \tab 298 \tab 76 \tab 4 \tab 37 \tab 2832 \cr
   2006 \tab 3492 \tab 52 \tab 226 \tab 240 \tab 372 \tab 124 \tab 2 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 53 \tab 208 \tab 259 \tab 318 \tab 96 \tab 1 \tab 5 \tab 2867 \cr
   Total \tab 8025 \tab 183 \tab 786 \tab 1013 \tab 1285 \tab 369 \tab 14 \tab 71 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}},
\code{\link{xhaustn}}
}
\concept{Work Orientation}
\keyword{variable}
