% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{rhmewrk}
\alias{rhmewrk}
\title{How much housework r does}
\source{
General Social Survey https://gss.norc.org
}
\description{
rhmewrk
}
\details{
Question We're interested in knowing how much of the work around your home is done by different people. We don't mean taking care of children - just regular work around the house, like cooking, grocery shopping and doing little repair jobs.
A. How much of this work do you do?
}
\section{Values}{

\itemize{
\item \code{1} all
\item \code{2} most
\item \code{3} about half
\item \code{4} some
\item \code{5} very little or none
\item \code{6} not married-living with partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1660/vshow}{GSS Data Explorer page for \code{rhmewrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all \tab most \tab about half \tab some \tab very little or none \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 479 \tab 312 \tab 419 \tab 176 \tab 64 \tab 2 \tab 8 \tab 2904 \cr
   Total \tab 1444 \tab 479 \tab 312 \tab 419 \tab 176 \tab 64 \tab 2 \tab 8 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{willwed1}},
\code{\link{willwed2}}
}
\concept{Gender}
\keyword{variable}
