% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{rhowfree}
\alias{rhowfree}
\title{How much freedom does r have}
\source{
General Social Survey https://gss.norc.org
}
\description{
rhowfree
}
\details{
Question And about you yourself?  Would you say right now that you have complete freedom, a great deal of freedom, a moderate degree of freedom, not much freedom, or no freedom at all?
}
\section{Values}{

\itemize{
\item \code{1} complete freedom
\item \code{2} a great deal of freedom
\item \code{3} a moderate degree of freedom
\item \code{4} not much freedom
\item \code{5} no freedom at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2664/vshow}{GSS Data Explorer page for \code{rhowfree}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab complete freedom \tab a great deal of freedom \tab a moderate degree of freedom \tab not much freedom \tab no freedom at all \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 238 \tab 769 \tab 333 \tab 53 \tab 5 \tab 7 \tab 14 \tab 2817 \cr
   Total \tab 1398 \tab 238 \tab 769 \tab 333 \tab 53 \tab 5 \tab 7 \tab 14 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{freenow}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{satdemoc}}
}
\concept{Freedom}
\keyword{variable}
