% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{rifles50}
\alias{rifles50}
\title{Should high power rifle sale be limited to military}
\source{
General Social Survey https://gss.norc.org
}
\description{
rifles50
}
\details{
Question Currently under federal law, very high power, 50-caliber rifles that can penetrate armor from a mile away are available to people on the same basis as standard hunting rifles. Should such very high power rifles be 1) restricted Only to the police and military, or 2) available to civilians like standard hunting rifles as they are at present?
}
\section{Values}{

\itemize{
\item \verb{[1]} restricted to the police and military
\item \verb{[2]} available to civilians like other hunting rifles
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3704/vshow}{GSS Data Explorer page for \code{rifles50}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab restricted to the police and military \tab available to civilians like other hunting rifles \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3139 \tab 1151 \tab 187 \tab 21 \tab 12 \tab 4510 \cr
   Total \tab 3139 \tab 1151 \tab 187 \tab 21 \tab 12 \tab 4510 \cr
}
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\keyword{variable}
