% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{rightpeo}
\alias{rightpeo}
\title{Religion helps people to meet the right kind of people}
\source{
General Social Survey https://gss.norc.org
}
\description{
rightpeo
}
\details{
Question Do you agree or disagree that…
D. ...that practicing a religion helps people to meet the right kind of people?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4715/vshow}{GSS Data Explorer page for \code{rightpeo}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 220 \tab 573 \tab 391 \tab 130 \tab 28 \tab 17 \tab 6 \tab 2023 \cr
   Total \tab 658 \tab 220 \tab 573 \tab 391 \tab 130 \tab 28 \tab 17 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{vistholy}}
}
\concept{ISSP Religion}
\keyword{variable}
