% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{rimmdisc}
\alias{rimmdisc}
\title{Chances illegal immigrants favored over family}
\source{
General Social Survey https://gss.norc.org
}
\description{
rimmdisc
}
\details{
Question B. What about immigrants?  Is it very likely, somewhat likely, somewhat unlikely, or very unlikely that you or anyone in your family won't get a job or promotion while an equally or less qualified immigrant employee receives one instead?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1648/vshow}{GSS Data Explorer page for \code{rimmdisc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 126 \tab 309 \tab 427 \tab 480 \tab 110 \tab 22 \tab 2992 \cr
   Total \tab 1518 \tab 126 \tab 309 \tab 427 \tab 480 \tab 110 \tab 22 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}}
}
\concept{Multiculturalism}
\keyword{variable}
