% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{rincblls}
\alias{rincblls}
\title{Income alone is enough}
\source{
General Social Survey https://gss.norc.org
}
\description{
rincblls
}
\details{
Question Do you feel that the income from your job alone is enough to meet your family's usual monthly expenses and bills?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2817/vshow}{GSS Data Explorer page for \code{rincblls}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 969 \tab 827 \tab 944 \tab 11 \tab 14 \tab - \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 827 \tab 881 \tab 8 \tab 18 \tab - \tab - \tab 4510 \cr
   2010 \tab 857 \tab 506 \tab 646 \tab 5 \tab 30 \tab - \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 613 \tab 627 \tab 2 \tab 8 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 713 \tab 693 \tab 7 \tab 35 \tab - \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 966 \tab 979 \tab 11 \tab 14 \tab 19 \tab - \tab 3544 \cr
   Total \tab 8345 \tab 4452 \tab 4770 \tab 44 \tab 119 \tab 19 \tab 0 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{chngtme}},
\code{\link{condemnd}},
\code{\link{cowrkhlp}},
\code{\link{cowrkint}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{fairearn}},
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{fringeok}},
\code{\link{handmove}},
\code{\link{haveinfo}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{hlpequip}},
\code{\link{hlthdays}},
\code{\link{hrsrelax}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{hyperten}},
\code{\link{jobfind1}},
\code{\link{jobsecok}},
\code{\link{knowschd}},
\code{\link{knowschd1}},
\code{\link{knowwhat}},
\code{\link{laidoff}},
\code{\link{learnnew}},
\code{\link{lotofsay}},
\code{\link{manvsemp}},
\code{\link{misswork}},
\code{\link{mntlhlth}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{myskills}},
\code{\link{ntwkhard}},
\code{\link{opdevel}},
\code{\link{overwork}},
\code{\link{painarms}},
\code{\link{partteam}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{prodctiv}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{satjob1}},
\code{\link{secondwk}},
\code{\link{setthngs}},
\code{\link{slpprblm}},
\code{\link{spvtrfair}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{suphelp}},
\code{\link{teamsafe}},
\code{\link{toofewwk}},
\code{\link{trainops}},
\code{\link{trdunion}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{usedup}},
\code{\link{usetech}},
\code{\link{waypaid}},
\code{\link{weight}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbonus}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkpraise}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksmooth}},
\code{\link{wktopsat}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrktime}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Income: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{earndes}},
\code{\link{earnmore}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{fairearn}},
\code{\link{famfinan}},
\code{\link{finalter}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{finrela}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{impfinan}},
\code{\link{incdef}},
\code{\link{inclstjb}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{justpay}},
\code{\link{minfood}},
\code{\link{minfour}},
\code{\link{mininc}},
\code{\link{minthree}},
\code{\link{othpay}},
\code{\link{owninc}},
\code{\link{pastpay}},
\code{\link{paychild}},
\code{\link{paychnge}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{payfam1}},
\code{\link{payhard}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{povline}},
\code{\link{realinc}},
\code{\link{realrinc}},
\code{\link{rincom06}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{salsatfn}},
\code{\link{satfin}},
\code{\link{welfare1}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}}
}
\concept{Income}
\concept{Quality of Working Life}
\keyword{variable}
