% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{rincom06}
\alias{rincom06}
\title{Respondents income}
\source{
General Social Survey https://gss.norc.org
}
\description{
rincom06
}
\details{
Question In which of these groups did your earnings from (OCC), from all sources for (last year) fall? That is, before taxes or other deductions. Just tell me the letter.
}
\section{Values}{

\itemize{
\item \code{1} under $1 000
\item \code{2} $1 000 to 2 999
\item \code{3} $3 000 to 3 999
\item \code{4} $4 000 to 4 999
\item \code{5} $5 000 to 5 999
\item \code{6} $6 000 to 6 999
\item \code{7} $7 000 to 7 999
\item \code{8} $8 000 to 9 999
\item \code{9} $10000 to 12499
\item \code{10} $12500 to 14999
\item \code{11} $15000 to 17499
\item \code{12} $17500 to 19999
\item \code{13} $20000 to 22499
\item \code{14} $22500 to 24999
\item \code{15} $25000 to 29999
\item \code{16} $30000 to 34999
\item \code{17} $35000 to 39999
\item \code{18} $40000 to 49999
\item \code{19} $50000 to 59999
\item \code{20} $60000 to 74999
\item \code{21} $75000 to $89999
\item \code{22} $90000 to $109999
\item \code{23} $110000 to $129999
\item \code{24} $130000 to $149999
\item \code{25} $150000 or over
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2008, 2010, 2012, 2014 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/118/vshow}{GSS Data Explorer page for \code{rincom06}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab under $1 000 \tab $1 000 to 2 999 \tab $3 000 to 3 999 \tab $4 000 to 4 999 \tab $5 000 to 5 999 \tab $6 000 to 6 999 \tab $7 000 to 7 999 \tab $8 000 to 9 999 \tab $10000 to 12499 \tab $12500 to 14999 \tab $15000 to 17499 \tab $17500 to 19999 \tab $20000 to 22499 \tab $22500 to 24999 \tab $25000 to 29999 \tab $30000 to 34999 \tab $35000 to 39999 \tab $40000 to 49999 \tab $50000 to 59999 \tab $60000 to 74999 \tab $75000 to $89999 \tab $90000 to $109999 \tab $110000 to $129999 \tab $130000 to $149999 \tab $150000 or over \tab don't know \tab no answer \tab refused \tab Total \cr
   2006 \tab 1472 \tab 53 \tab 75 \tab 47 \tab 53 \tab 41 \tab 42 \tab 37 \tab 61 \tab 123 \tab 108 \tab 110 \tab 111 \tab 139 \tab 109 \tab 193 \tab 222 \tab 169 \tab 301 \tab 198 \tab 176 \tab 104 \tab 64 \tab 50 \tab 22 \tab 61 \tab 71 \tab 53 \tab 245 \tab 4510 \cr
   2008 \tab 693 \tab 22 \tab 39 \tab 27 \tab 13 \tab 19 \tab 16 \tab 20 \tab 24 \tab 47 \tab 63 \tab 40 \tab 42 \tab 44 \tab 76 \tab 70 \tab 94 \tab 71 \tab 125 \tab 85 \tab 87 \tab 61 \tab 38 \tab 28 \tab 6 \tab 32 \tab 16 \tab 40 \tab 85 \tab 2023 \cr
   2010 \tab 684 \tab 29 \tab 43 \tab 33 \tab 23 \tab 26 \tab 28 \tab 19 \tab 34 \tab 72 \tab 38 \tab 51 \tab 37 \tab 56 \tab 51 \tab 66 \tab 78 \tab 67 \tab 102 \tab 88 \tab 98 \tab 63 \tab 36 \tab 28 \tab 14 \tab 22 \tab 13 \tab - \tab 145 \tab 2044 \cr
   2012 \tab 723 \tab 27 \tab 32 \tab 30 \tab 22 \tab 20 \tab 24 \tab 22 \tab 23 \tab 54 \tab 49 \tab 44 \tab 45 \tab 51 \tab 54 \tab 78 \tab 71 \tab 62 \tab 121 \tab 78 \tab 84 \tab 51 \tab 41 \tab 21 \tab 7 \tab 35 \tab 28 \tab - \tab 77 \tab 1974 \cr
   2014 \tab 923 \tab 32 \tab 36 \tab 38 \tab 25 \tab 32 \tab 24 \tab 14 \tab 30 \tab 56 \tab 55 \tab 48 \tab 39 \tab 76 \tab 67 \tab 90 \tab 112 \tab 79 \tab 154 \tab 145 \tab 114 \tab 88 \tab 53 \tab 37 \tab 21 \tab 58 \tab 19 \tab - \tab 73 \tab 2538 \cr
   Total \tab 4495 \tab 163 \tab 225 \tab 175 \tab 136 \tab 138 \tab 134 \tab 112 \tab 172 \tab 352 \tab 313 \tab 293 \tab 274 \tab 366 \tab 357 \tab 497 \tab 577 \tab 448 \tab 803 \tab 594 \tab 559 \tab 367 \tab 232 \tab 164 \tab 70 \tab 208 \tab 147 \tab 93 \tab 625 \tab 13089 \cr
}
}

\seealso{
Other Core: 
\code{\link{dotdata}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{region}},
\code{\link{rincom16}},
\code{\link{rincom98}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Core}
\keyword{variable}
