% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{rincom16}
\alias{rincom16}
\title{Respondents income}
\source{
General Social Survey https://gss.norc.org
}
\description{
rincom16
}
\details{
Question In which of these groups did your earnings from (job), from all sources for 2015 fall?  That is, before taxes or other deductions.  Just tell me the letter.
}
\section{Values}{

\itemize{
\item \code{1} under $1,000
\item \code{2} $1,000 to $2,999
\item \code{3} $3,000 to $3,999
\item \code{4} $4,000 to $4,999
\item \code{5} $5,000 to $5,999
\item \code{6} $6,000 to $6,999
\item \code{7} $7,000 to $7,999
\item \code{8} $8,000 to $9,999
\item \code{9} $10,000 to $12,499
\item \code{10} $12,500 to $14,999
\item \code{11} $15,000 to $17,499
\item \code{12} $17,500 to $19,999
\item \code{13} $20,000 to $22,499
\item \code{14} $22,500 to $24,999
\item \code{15} $25,000 to $29,999
\item \code{16} $30,000 to $34,999
\item \code{17} $35,000 to $39,999
\item \code{18} $40,000 to $49,999
\item \code{19} $50,000 to $59,999
\item \code{20} $60,000 to $74,999
\item \code{21} $75,000 to $89,999
\item \code{22} $90,000 to $109,999
\item \code{23} $110,000 to $129,999
\item \code{24} $130,000 to $149,999
\item \code{25} $150,000 to $169,999
\item \code{26} $170,000 or over
\item \code{27} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6168/vshow}{GSS Data Explorer page for \code{rincom16}}.
}

\seealso{
Other Core: 
\code{\link{dotdata}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{region}},
\code{\link{rincom06}},
\code{\link{rincom98}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Core}
\keyword{variable}
