% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{rintjob}
\alias{rintjob}
\title{R's job is interesting}
\source{
General Social Survey https://gss.norc.org
}
\description{
rintjob
}
\details{
Question For each statement about your main job below, please circle one code to show how much you agree or disagree that it applies to your job.
E. My job is interesting
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4545/vshow}{GSS Data Explorer page for \code{rintjob}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 224 \tab 442 \tab 112 \tab 69 \tab 24 \tab 2 \tab 25 \tab 1537 \cr
   1998 \tab 1967 \tab 188 \tab 423 \tab 128 \tab 60 \tab 27 \tab 3 \tab 36 \tab 2832 \cr
   2006 \tab 3492 \tab 342 \tab 524 \tab 65 \tab 59 \tab 25 \tab 1 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 325 \tab 410 \tab 110 \tab 64 \tab 24 \tab 1 \tab 6 \tab 2867 \cr
   Total \tab 8025 \tab 1079 \tab 1799 \tab 415 \tab 252 \tab 100 \tab 7 \tab 69 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rleisure}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}},
\code{\link{xhaustn}}
}
\concept{Work Orientation}
\keyword{variable}
