% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{rleisure}
\alias{rleisure}
\title{R's job leaves much leisure time}
\source{
General Social Survey https://gss.norc.org
}
\description{
rleisure
}
\details{
Question For each statement about your main job below, please circle one code to show how much you agree or disagree that it applies to your job.
D. My job leaves a lot of leisure time
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4544/vshow}{GSS Data Explorer page for \code{rleisure}}.
}

\seealso{
Other Work Orientation: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}},
\code{\link{xhaustn}}

Other Job: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}},
\code{\link{xhaustn}}
}
\concept{Job}
\concept{Work Orientation}
\keyword{variable}
