% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_611.R
\name{rlooks}
\alias{rlooks}
\title{R physical attractiveness rating}
\source{
General Social Survey https://gss.norc.org
}
\description{
rlooks
}
\details{
Question How physically attractive is the respondent?
}
\section{Values}{

\itemize{
\item \code{1} very unattractive
\item \code{2} unattractive
\item \code{3} about average
\item \code{4} attractive
\item \code{5} very attractive
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6659/vshow}{GSS Data Explorer page for \code{rlooks}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very unattractive \tab unattractive \tab about average \tab attractive \tab very attractive \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab - \tab 48 \tab 140 \tab 1483 \tab 797 \tab 183 \tab 9 \tab 207 \tab - \tab 2867 \cr
   2018 \tab - \tab 47 \tab 137 \tab 1129 \tab 608 \tab 121 \tab 11 \tab 295 \tab - \tab 2348 \cr
   2022 \tab 1864 \tab 31 \tab 119 \tab 915 \tab 485 \tab 122 \tab 8 \tab - \tab - \tab 3544 \cr
   Total \tab 1864 \tab 126 \tab 396 \tab 3527 \tab 1890 \tab 426 \tab 28 \tab 502 \tab 0 \tab 8759 \cr
}
}

\keyword{variable}
