% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{rnrltu18}
\alias{rnrltu18}
\title{R is not related to child under 18 in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
rnrltu18
}
\details{
Question Are there any individuals under the age of 18 in this household who are not related to you in any way? For example, the child of a roommate or lodger.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8026/vshow}{GSS Data Explorer page for \code{rnrltu18}}.
}

\keyword{variable}
