% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{rntrltd18}
\alias{rntrltd18}
\title{R is not related to another adult in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
rntrltd18
}
\details{
Question Is there anyone in your household who you are not related to? For example, an adult roommate or lodger?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8027/vshow}{GSS Data Explorer page for \code{rntrltd18}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3265 \tab 70 \tab 202 \tab 1 \tab 1 \tab 5 \tab 3544 \cr
   Total \tab 3265 \tab 70 \tab 202 \tab 1 \tab 1 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
