% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{robbry}
\alias{robbry}
\title{Forcefully robbed during the last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
robbry
}
\details{
Question During the last year, did anyone take something directly from you by using force--such as a stickup, mugging, or threat?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/678/vshow}{GSS Data Explorer page for \code{robbry}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1973 \tab - \tab 24 \tab 1476 \tab 4 \tab 1504 \cr
   1974 \tab - \tab 53 \tab 1420 \tab 11 \tab 1484 \cr
   1976 \tab - \tab 24 \tab 1467 \tab 8 \tab 1499 \cr
   1977 \tab - \tab 29 \tab 1496 \tab 5 \tab 1530 \cr
   1980 \tab - \tab 30 \tab 1436 \tab 2 \tab 1468 \cr
   1982 \tab - \tab 52 \tab 1804 \tab 4 \tab 1860 \cr
   1984 \tab - \tab 23 \tab 1448 \tab 2 \tab 1473 \cr
   1985 \tab - \tab 31 \tab 1500 \tab 3 \tab 1534 \cr
   1987 \tab - \tab 44 \tab 1769 \tab 6 \tab 1819 \cr
   1988 \tab 504 \tab 15 \tab 961 \tab 1 \tab 1481 \cr
   1989 \tab 504 \tab 13 \tab 1013 \tab 7 \tab 1537 \cr
   1990 \tab 455 \tab 13 \tab 901 \tab 3 \tab 1372 \cr
   1991 \tab 524 \tab 17 \tab 970 \tab 6 \tab 1517 \cr
   1993 \tab 531 \tab 28 \tab 1045 \tab 2 \tab 1606 \cr
   1994 \tab 1767 \tab 31 \tab 1192 \tab 2 \tab 2992 \cr
   Total \tab 4285 \tab 427 \tab 19898 \tab 66 \tab 24676 \cr
}
}

\seealso{
Other Core: 
\code{\link{burglr}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{owngun}},
\code{\link{pistol}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{rifle}},
\code{\link{shotgun}}
}
\concept{Core}
\keyword{variable}
