% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{robject}
\alias{robject}
\title{Did r object to or delay answering card c4}
\source{
General Social Survey https://gss.norc.org
}
\description{
robject
}
\details{
Question INTERVIEWER: DID THE RESPONDENT OBJECT TO OR DELAY ANSWERING IMMSTATS?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, objected verbally
\item \verb{[2]} yes, delayed, kept silent for a while
\item \verb{[3]} no
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3029/vshow}{GSS Data Explorer page for \code{robject}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab yes, objected verbally \tab yes, delayed, kept silent for a while \tab no \tab other \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2004 \tab 2540 \tab 5 \tab 4 \tab 248 \tab 3 \tab 7 \tab 5 \tab - \tab 2812 \cr
   Total \tab 2540 \tab 5 \tab 4 \tab 248 \tab 3 \tab 7 \tab 5 \tab 0 \tab 2812 \cr
}
}

\seealso{
Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{typevisa}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Interview: 
\code{\link{comprend}},
\code{\link{consent}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{easyget}},
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{id}},
\code{\link{intage}},
\code{\link{intage_0418}},
\code{\link{intethn}},
\code{\link{inthisp}},
\code{\link{intid}},
\code{\link{intrace1}},
\code{\link{intrace2}},
\code{\link{intrace3}},
\code{\link{intrwght}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{intyrs_0418}},
\code{\link{lngthinv}},
\code{\link{mode}()},
\code{\link{rundrstd}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{sciintro}},
\code{\link{scimode}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{year}()}
}
\concept{Immigration}
\concept{Interview}
\keyword{variable}
