% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{romance}
\alias{romance}
\title{Do you have a main romantic involvement}
\source{
General Social Survey https://gss.norc.org
}
\description{
romance
}
\details{
Question Do you have a main romantic involvement - a (man/woman) you think of as a steady, a lover, a partner, or whatever?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} has same gender partner
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1649/vshow}{GSS Data Explorer page for \code{romance}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab has same gender partner \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 2126 \tab 319 \tab 427 \tab 1 \tab 2 \tab 29 \tab - \tab 2904 \cr
   1998 \tab 2109 \tab 272 \tab 415 \tab 3 \tab 3 \tab 30 \tab - \tab 2832 \cr
   Total \tab 4235 \tab 591 \tab 842 \tab 4 \tab 5 \tab 59 \tab 0 \tab 5736 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{balwkfam}},
\code{\link{children}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{earnmore}},
\code{\link{emoteoth}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{febear}},
\code{\link{fehire}},
\code{\link{feimp}},
\code{\link{fejobaff}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{flextime}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{haprom}},
\code{\link{hmemaker}},
\code{\link{livewith}},
\code{\link{manprof}},
\code{\link{marfree}},
\code{\link{mebear}},
\code{\link{men}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{parleave}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{rhmewrk}},
\code{\link{rushed}},
\code{\link{selffrst}},
\code{\link{sexhar}},
\code{\link{sexpromo}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{twoincs}},
\code{\link{willwed1}},
\code{\link{willwed2}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{wrkclass}},
\code{\link{yourself}}

Other Romantic Relationships: 
\code{\link{hapcohab}},
\code{\link{haprom}},
\code{\link{livewith}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{relgendr}},
\code{\link{willwed1}},
\code{\link{willwed2}}
}
\concept{Gender}
\concept{Romantic Relationships}
\keyword{variable}
