% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{rprtst1}
\alias{rprtst1}
\title{Would you attend a public meeting}
\source{
General Social Survey https://gss.norc.org
}
\description{
rprtst1
}
\details{
Question Would you or would you not do any of the following to protest against a government action you strongly opposed? please CHECK ONE BOX ON EACH LINE.
A. Attend a public meeting organized to protest against the government
}
\section{Values}{

\itemize{
\item \code{1} definietly would
\item \code{2} probably would
\item \code{3} probably would not
\item \code{4} definitely would not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4090/vshow}{GSS Data Explorer page for \code{rprtst1}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}

Other Protest: 
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}}
}
\concept{ISSP Role of Gov't}
\concept{Protest}
\keyword{variable}
