% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{rprtst15}
\alias{rprtst15}
\title{How many times be a public meeting past 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
rprtst15
}
\details{
Question And in the past five years how many times have you done each of the following to protest against a government action you strongly oppose? PLEASE CHECK ONE BOX ON EACH LINE
A. Attended a public meeting organized to protest against the government.
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} once
\item \code{3} more than once
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4092/vshow}{GSS Data Explorer page for \code{rprtst15}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab never \tab once \tab more than once \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 1127 \tab 118 \tab 59 \tab 28 \tab 2904 \cr
   Total \tab 1572 \tab 1127 \tab 118 \tab 59 \tab 28 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
