% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{rprtst35}
\alias{rprtst35}
\title{How many times gone on a protest march past 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
rprtst35
}
\details{
Question And in the past five years how many times have you done each of the following to protest against a government action you strongly oppose? PLEASE CHECK ONE BOX ON EACH LINE
B. Gone on a protest march or demonstration.
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} once
\item \verb{[3]} more than once
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4093/vshow}{GSS Data Explorer page for \code{rprtst35}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab never \tab once \tab more than once \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 1177 \tab 87 \tab 31 \tab 1 \tab 36 \tab 2904 \cr
   Total \tab 1572 \tab 1177 \tab 87 \tab 31 \tab 1 \tab 36 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
