% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{rptcowrk}
\alias{rptcowrk}
\title{Report coworkers threatening secrecy of work}
\source{
General Social Survey https://gss.norc.org
}
\description{
rptcowrk
}
\details{
Question When faced with a conflict between loyalty to one''s employer and loyalty to a co-worker who is violating rules for protecting SECRET and TOP SECRET information, a person should:
}
\section{Values}{

\itemize{
\item \verb{[1]} report coworker
\item \verb{[2]} ask coworker to stop
\item \verb{[3]} report cowrkr if behav cont
\item \verb{[4]} not get involved
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5207/vshow}{GSS Data Explorer page for \code{rptcowrk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab report coworker \tab ask coworker to stop \tab report cowrkr if behav cont \tab not get involved \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 591 \tab 85 \tab 606 \tab 124 \tab 61 \tab 7 \tab 2992 \cr
   Total \tab 1518 \tab 591 \tab 85 \tab 606 \tab 124 \tab 61 \tab 7 \tab 2992 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{secdocs}},
\code{\link{sectech}}
}
\concept{National Security}
\keyword{variable}
