% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{rptprobs}
\alias{rptprobs}
\title{Ppl feel free to report problems in workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
rptprobs
}
\details{
Question B. People feel free to report problems in my workplace without fear of negative consequences.
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2999/vshow}{GSS Data Explorer page for \code{rptprobs}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 948 \tab 511 \tab 195 \tab 140 \tab 31 \tab 42 \tab 2812 \cr
   Total \tab 945 \tab 948 \tab 511 \tab 195 \tab 140 \tab 31 \tab 42 \tab 2812 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{difstand}},
\code{\link{lackinfo}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{putdown}}
}
\concept{Workplace Conflict}
\keyword{variable}
