% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_666.R
\name{rsblng18}
\alias{rsblng18}
\title{R has sibling in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
rsblng18
}
\details{
Question Are you the sibling of another individual 18 or older in your house?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8030/vshow}{GSS Data Explorer page for \code{rsblng18}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3243 \tab 64 \tab 231 \tab 1 \tab 5 \tab 3544 \cr
   Total \tab 3243 \tab 64 \tab 231 \tab 1 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
