% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_499.R
\name{rsecjob}
\alias{rsecjob}
\title{R's job is secure}
\source{
General Social Survey https://gss.norc.org
}
\description{
rsecjob
}
\details{
Question For each statement about your main job below, please circle one code to show how much you agree or disagree that it applies to your job.
A. My job is secure
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4541/vshow}{GSS Data Explorer page for \code{rsecjob}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 248 \tab 409 \tab 116 \tab 74 \tab 21 \tab 6 \tab 24 \tab 1537 \cr
   1998 \tab 1967 \tab 197 \tab 377 \tab 127 \tab 96 \tab 24 \tab 7 \tab 37 \tab 2832 \cr
   2006 \tab 3492 \tab 288 \tab 431 \tab 129 \tab 137 \tab 25 \tab 6 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 316 \tab 392 \tab 121 \tab 84 \tab 21 \tab 1 \tab 5 \tab 2867 \cr
   Total \tab 8025 \tab 1049 \tab 1609 \tab 493 \tab 391 \tab 91 \tab 20 \tab 68 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{hrsmoney}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
