% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{rspctrel}
\alias{rspctrel}
\title{We must respect all religions}
\source{
General Social Survey https://gss.norc.org
}
\description{
rspctrel
}
\details{
Question How much do you agree or disagree with the following statements?
B. We must respect all religions. Would you say…
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4700/vshow}{GSS Data Explorer page for \code{rspctrel}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 433 \tab 666 \tab 116 \tab 106 \tab 31 \tab 8 \tab 5 \tab 2023 \cr
   Total \tab 658 \tab 433 \tab 666 \tab 116 \tab 106 \tab 31 \tab 8 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{ISSP Religion}
\keyword{variable}
