% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{rstpprnt}
\alias{rstpprnt}
\title{R is step-parent of child under 18 in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
rstpprnt
}
\details{
Question Are you the stepparent of at least one individual under the age of 18 in this household?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8031/vshow}{GSS Data Explorer page for \code{rstpprnt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3189 \tab 30 \tab 318 \tab 1 \tab 6 \tab 3544 \cr
   Total \tab 3189 \tab 30 \tab 318 \tab 1 \tab 6 \tab 3544 \cr
}
}

\keyword{variable}
