% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_666.R
\name{rstpprnt18}
\alias{rstpprnt18}
\title{R is step-parent of another adult in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
rstpprnt18
}
\details{
Question To clarify, are you the stepparent of at least one individual 18 or older in your household
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8032/vshow}{GSS Data Explorer page for \code{rstpprnt18}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2022 \tab 3440 \tab 6 \tab 98 \tab 3544 \cr
   Total \tab 3440 \tab 6 \tab 98 \tab 3544 \cr
}
}

\keyword{variable}
