% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{runhosp}
\alias{runhosp}
\title{Hospitals should mainly be run by private}
\source{
General Social Survey https://gss.norc.org
}
\description{
runhosp
}
\details{
Question Do you think each of the following should mainly be run by private organizations or companies, or by the government? PLEASE CHECK ONE BOX ON EACH LINE.
B. Hospitals
}
\section{Values}{

\itemize{
\item \verb{[1]} by private
\item \verb{[2]} by the government
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4105/vshow}{GSS Data Explorer page for \code{runhosp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab by private \tab by the government \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 853 \tab 303 \tab 139 \tab 37 \tab 2904 \cr
   Total \tab 1572 \tab 853 \tab 303 \tab 139 \tab 37 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
