% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{rushed}
\alias{rushed}
\title{How often r feels rushed}
\source{
General Social Survey https://gss.norc.org
}
\description{
rushed
}
\details{
Question In general, how do you feel about your time--would you say you always feel rushed even to do things you have to do, Only sometimes feel rushed, or almost never feel rushed.
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} sometimes
\item \verb{[3]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   B/C/- \tab 2004 \cr
   C/-/- \tab 2016 \cr
   Full \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/774/vshow}{GSS Data Explorer page for \code{rushed}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab always \tab sometimes \tab never \tab no answer \tab don't know \tab Total \cr
   1982 \tab - \tab 442 \tab 967 \tab 443 \tab 8 \tab - \tab 1860 \cr
   1996 \tab 1444 \tab 434 \tab 757 \tab 261 \tab 7 \tab 1 \tab 2904 \cr
   2004 \tab 1828 \tab 304 \tab 513 \tab 160 \tab 6 \tab 1 \tab 2812 \cr
   2016 \tab 1888 \tab 215 \tab 542 \tab 218 \tab 4 \tab - \tab 2867 \cr
   Total \tab 5160 \tab 1395 \tab 2779 \tab 1082 \tab 25 \tab 2 \tab 10443 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{stresswk}}
}
\concept{Core}
\concept{Gender}
\keyword{variable}
