% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_611.R
\name{rweight}
\alias{rweight}
\title{R weight rating}
\source{
General Social Survey https://gss.norc.org
}
\description{
rweight
}
\details{
Question How would you rate the respondent's weight?
}
\section{Values}{

\itemize{
\item \code{1} very overweight
\item \code{2} slightly overweight
\item \code{3} about the right weight
\item \code{4} slightly underweight
\item \code{5} very underweight
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6661/vshow}{GSS Data Explorer page for \code{rweight}}.
}

\keyword{variable}
