% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{rwrkbaby}
\alias{rwrkbaby}
\title{Did r work when kid was preschooler}
\source{
General Social Survey https://gss.norc.org
}
\description{
rwrkbaby
}
\details{
Question Did you work outside the home full-time, part-time, or not at all...
B. And what about when a child was under school age?
}
\section{Values}{

\itemize{
\item \code{1} yes, full-time
\item \code{2} yes,part-time
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4397/vshow}{GSS Data Explorer page for \code{rwrkbaby}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, full-time \tab yes,part-time \tab no \tab don't know \tab no answer \tab Total \cr
   1988 \tab 1148 \tab 70 \tab 67 \tab 174 \tab 5 \tab 17 \tab 1481 \cr
   1994 \tab 1545 \tab 586 \tab 136 \tab 281 \tab 72 \tab 372 \tab 2992 \cr
   2002 \tab 1936 \tab 482 \tab 121 \tab 172 \tab 42 \tab 12 \tab 2765 \cr
   Total \tab 4629 \tab 1138 \tab 324 \tab 627 \tab 119 \tab 401 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabit}},
\code{\link{divhubby}},
\code{\link{earnsmor}},
\code{\link{evdiv}},
\code{\link{mawork14}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
