% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{rwrkgrwn}
\alias{rwrkgrwn}
\title{Did r work after kids left home}
\source{
General Social Survey https://gss.norc.org
}
\description{
rwrkgrwn
}
\details{
Question Did you work outside the home full-time, part-time, or not at all...
D. And how about after the children left home?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, full-time
\item \verb{[2]} yes,part-time
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4399/vshow}{GSS Data Explorer page for \code{rwrkgrwn}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, full-time \tab yes,part-time \tab no \tab don't know \tab no answer \tab Total \cr
   1988 \tab 1148 \tab 103 \tab 31 \tab 40 \tab 137 \tab 22 \tab 1481 \cr
   1994 \tab 1545 \tab 475 \tab 68 \tab 76 \tab 411 \tab 417 \tab 2992 \cr
   2002 \tab 1936 \tab 367 \tab 56 \tab 36 \tab 358 \tab 12 \tab 2765 \cr
   Total \tab 4629 \tab 945 \tab 155 \tab 152 \tab 906 \tab 451 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabit}},
\code{\link{divhubby}},
\code{\link{earnsmor}},
\code{\link{evdiv}},
\code{\link{mawork14}},
\code{\link{rwrkbaby}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
