% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{rwrksch}
\alias{rwrksch}
\title{Did r work after youngest in school}
\source{
General Social Survey https://gss.norc.org
}
\description{
rwrksch
}
\details{
Question Did you work outside the home full-time, part-time, or not at all...
C. After the youngest child started school?
}
\section{Values}{

\itemize{
\item \code{1} yes, full-time
\item \code{2} yes,part-time
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4398/vshow}{GSS Data Explorer page for \code{rwrksch}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, full-time \tab yes,part-time \tab no \tab don't know \tab no answer \tab Total \cr
   1988 \tab 1148 \tab 103 \tab 69 \tab 87 \tab 55 \tab 19 \tab 1481 \cr
   1994 \tab 1545 \tab 619 \tab 131 \tab 128 \tab 189 \tab 380 \tab 2992 \cr
   2002 \tab 1936 \tab 481 \tab 118 \tab 88 \tab 130 \tab 12 \tab 2765 \cr
   Total \tab 4629 \tab 1203 \tab 318 \tab 303 \tab 374 \tab 411 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabit}},
\code{\link{divhubby}},
\code{\link{earnsmor}},
\code{\link{evdiv}},
\code{\link{mawork14}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
