% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_499.R
\name{ryrsfirm}
\alias{ryrsfirm}
\title{Importance of tenure}
\source{
General Social Survey https://gss.norc.org
}
\description{
ryrsfirm
}
\details{
Question Think of two people doing the same kind of work in your place. What Do you personally think is important in deciding how much to pay them?
Looking at the things below, please write ''1'' in the box next to the thing that you think is most important at your place of work. Then write ''2'' next to the thing that you think is next most important, and ''3'' next to the thing that you think is third most important. Leave the other boxes blank.
At my work, deciding on two people doing the same kind of work, how important is...
A. How long the employee has been with the firm
}
\section{Values}{

\itemize{
\item \verb{[1]} most impt
\item \verb{[2]} 2nd most impt
\item \verb{[3]} 3rd most impt
\item \verb{[4]} not chosen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4533/vshow}{GSS Data Explorer page for \code{ryrsfirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most impt \tab 2nd most impt \tab 3rd most impt \tab not chosen \tab don't know \tab no answer \tab Total \cr
   1989 \tab 639 \tab 74 \tab 132 \tab 226 \tab 263 \tab 25 \tab 178 \tab 1537 \cr
   Total \tab 639 \tab 74 \tab 132 \tab 226 \tab 263 \tab 25 \tab 178 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{hrsmoney}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rsecjob}}
}
\concept{Work Orientation}
\keyword{variable}
