% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{safefrst}
\alias{safefrst}
\title{No shortcuts on worker safety}
\source{
General Social Survey https://gss.norc.org
}
\description{
safefrst
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
L. There are no significant compromises or shortcuts taken when worker safety is at stake
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2785/vshow}{GSS Data Explorer page for \code{safefrst}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 733 \tab 815 \tab 157 \tab 47 \tab 26 \tab 18 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 710 \tab 790 \tab 146 \tab 30 \tab 41 \tab 17 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 464 \tab 583 \tab 84 \tab 16 \tab 12 \tab 28 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 542 \tab 587 \tab 84 \tab 13 \tab 14 \tab 10 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 542 \tab 725 \tab 119 \tab 13 \tab 14 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 834 \tab 869 \tab 197 \tab 36 \tab 29 \tab 13 \tab 20 \tab 3544 \cr
   Total \tab 8336 \tab 3825 \tab 4369 \tab 787 \tab 155 \tab 136 \tab 121 \tab 20 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
