% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{safetywk}
\alias{safetywk}
\title{Worker safety priority at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
safetywk
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
K. The safety of workers is a high priority with management where I work
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2784/vshow}{GSS Data Explorer page for \code{safetywk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 770 \tab 815 \tab 136 \tab 37 \tab 21 \tab 17 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 758 \tab 765 \tab 136 \tab 23 \tab 34 \tab 18 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 477 \tab 562 \tab 90 \tab 21 \tab 10 \tab 27 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 547 \tab 602 \tab 73 \tab 12 \tab 8 \tab 8 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 547 \tab 734 \tab 106 \tab 18 \tab 9 \tab 34 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 766 \tab 966 \tab 174 \tab 45 \tab 14 \tab 14 \tab 19 \tab 3544 \cr
   Total \tab 8336 \tab 3865 \tab 4444 \tab 715 \tab 156 \tab 96 \tab 118 \tab 19 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{teamsafe}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
