% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{salfergt}
\alias{salfergt}
\title{Salience, women's rights}
\source{
General Social Survey https://gss.norc.org
}
\description{
salfergt
}
\details{
Question How often would you say that you and your friends think about topics we've been discussing during the interview? would you say that you and your friends think about (READ EACH ITEM A-E) very often, sometimes, or almost never?
A. Women's rights
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} sometimes
\item \verb{[3]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/732/vshow}{GSS Data Explorer page for \code{salfergt}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very often \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   1978 \tab - \tab 284 \tab 743 \tab 495 \tab 5 \tab 5 \tab 1532 \cr
   Total \tab 0 \tab 284 \tab 743 \tab 495 \tab 5 \tab 5 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{unemp5}}
}
\concept{Core}
\keyword{variable}
