% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{sallabor}
\alias{sallabor}
\title{Salience, labor leaders}
\source{
General Social Survey https://gss.norc.org
}
\description{
sallabor
}
\details{
Question How often would you say that you and your friends think about topics we've been discussing during the interview? would you say that you and your friends think about (READ EACH ITEM A-E) very often, sometimes, or almost never?
B. The people running organized labor
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} sometimes
\item \code{3} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/733/vshow}{GSS Data Explorer page for \code{sallabor}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very often \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   1978 \tab - \tab 270 \tab 659 \tab 587 \tab 10 \tab 6 \tab 1532 \cr
   Total \tab 0 \tab 270 \tab 659 \tab 587 \tab 10 \tab 6 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{unemp5}}
}
\concept{Core}
\keyword{variable}
