% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{salsci}
\alias{salsci}
\title{Salience, scientific community}
\source{
General Social Survey https://gss.norc.org
}
\description{
salsci
}
\details{
Question How often would you say that you and your friends think about topics we've been discussing during the interview? would you say that you and your friends think about (READ EACH ITEM A-E) very often, sometimes, or almost never?
E. The scientific community
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} sometimes
\item \verb{[3]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/736/vshow}{GSS Data Explorer page for \code{salsci}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab sometimes \tab almost never \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1978 \tab - \tab 148 \tab 561 \tab 785 \tab 31 \tab 7 \tab - \tab 1532 \cr
   Total \tab 0 \tab 148 \tab 561 \tab 785 \tab 31 \tab 7 \tab 0 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abhave3}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{ablegal}},
\code{\link{abnomore}},
\code{\link{aborct}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{abspno}},
\code{\link{adoptus}},
\code{\link{adults}},
\code{\link{affrmact}},
\code{\link{age}},
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{agewed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{antiwar}},
\code{\link{arrest}},
\code{\link{attend}},
\code{\link{babies}},
\code{\link{babies_7222}},
\code{\link{babies_exp}},
\code{\link{babies_exp_22}},
\code{\link{barate}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blksimp}},
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{bored}},
\code{\link{born}},
\code{\link{brass}},
\code{\link{broadband}},
\code{\link{burglr}},
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{capfirm}},
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{cathtemp}},
\code{\link{childs}},
\code{\link{childs_exp}},
\code{\link{childs_exp_22}},
\code{\link{childsinhh}},
\code{\link{childsinhh_2122}},
\code{\link{chlddth}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldsoon}},
\code{\link{cigweek}},
\code{\link{civic}},
\code{\link{civright}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{cobarate}},
\code{\link{codeg}},
\code{\link{coden}},
\code{\link{coden_1222}},
\code{\link{codipged}},
\code{\link{coeduc}},
\code{\link{coeftotlt}},
\code{\link{coevwork}},
\code{\link{cofund}},
\code{\link{cogradtounder}},
\code{\link{cohrs1}},
\code{\link{cohrs2}},
\code{\link{coind10}},
\code{\link{cojail}},
\code{\link{cojew}},
\code{\link{colath}},
\code{\link{colcom}},
\code{\link{colhomo}},
\code{\link{colmil}},
\code{\link{colmslm}},
\code{\link{color}},
\code{\link{colrac}},
\code{\link{colsoc}},
\code{\link{commute}},
\code{\link{comprend}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlabory}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conorelgsp}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{contemp}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{convictd}},
\code{\link{coocc10}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{coother}},
\code{\link{copres10}},
\code{\link{copunish}},
\code{\link{corel}},
\code{\link{cosector}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{defwrkev}},
\code{\link{defwrknw}},
\code{\link{degree}},
\code{\link{dejavu}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{dipged}},
\code{\link{dipged_1222}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{divorce}},
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{dotdata}},
\code{\link{dotged}},
\code{\link{dotpeop}},
\code{\link{dotpres}},
\code{\link{dotsvp}},
\code{\link{dotthng}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{earnrs}},
\code{\link{educ}},
\code{\link{educop}},
\code{\link{eftotlt}},
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{era}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{esp}},
\code{\link{evsmoke}},
\code{\link{evwork}},
\code{\link{fair}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{family16sex}},
\code{\link{father}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{febear}},
\code{\link{febrass}},
\code{\link{fecare}},
\code{\link{fechld}},
\code{\link{feclergy}},
\code{\link{fedirty}},
\code{\link{fefam}},
\code{\link{fefight}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{fehelp}},
\code{\link{fehlpmil}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{femilop}},
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{fenumok}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fepresch}},
\code{\link{fepriest}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fework}},
\code{\link{feworkif}},
\code{\link{fewrite}},
\code{\link{fightair}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{form}},
\code{\link{found}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govcare}},
\code{\link{grace}},
\code{\link{gradtounder}},
\code{\link{granborn}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{guerilla}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}},
\code{\link{gunner}},
\code{\link{gunnum}},
\code{\link{hapcohab}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{healer}},
\code{\link{health}},
\code{\link{helpblk}},
\code{\link{helpful}},
\code{\link{helpnot}},
\code{\link{helpoth}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitnum}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{hompop}},
\code{\link{hompop_7222}},
\code{\link{hompop_exp}},
\code{\link{hompop_exp_22}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{id}},
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{if96who}},
\code{\link{immcrime}},
\code{\link{immjobs}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imppol}},
\code{\link{imprelax}},
\code{\link{impwork}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income16}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{indstatus}},
\code{\link{indtag}},
\code{\link{indus07}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{jewaj}},
\code{\link{jewtemp}},
\code{\link{jobaccmp}},
\code{\link{jobfind}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{jobtrain}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kidssol}},
\code{\link{king}},
\code{\link{landwar}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{lessnuke}},
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{libath}},
\code{\link{libcom}},
\code{\link{liberatr}},
\code{\link{libhomo}},
\code{\link{libmil}},
\code{\link{libmslm}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{libtemp}},
\code{\link{life}},
\code{\link{likediff}},
\code{\link{lockedup}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madeath}},
\code{\link{madeg}},
\code{\link{maeduc}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maindstatus}},
\code{\link{maindtag}},
\code{\link{major1}},
\code{\link{major1_1222}},
\code{\link{major2}},
\code{\link{major2_1222}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{maoccindv}},
\code{\link{maoccstatus}},
\code{\link{maocctag}},
\code{\link{mapa}},
\code{\link{mapres10}},
\code{\link{mapres105plus}},
\code{\link{mapres80}},
\code{\link{marital}},
\code{\link{marsame}},
\code{\link{marsame1}},
\code{\link{marsamey}},
\code{\link{martype}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{mawrkslf}},
\code{\link{mechanic}},
\code{\link{meserve}},
\code{\link{migage}},
\code{\link{mighist}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{mindbody}},
\code{\link{mininc}},
\code{\link{minmilop}},
\code{\link{mobile16}},
\code{\link{mobiledata}},
\code{\link{morenuke}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{neargod}},
\code{\link{news}()},
\code{\link{nonuke}},
\code{\link{norelgsp16}},
\code{\link{nukewar}},
\code{\link{nurse}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occindv}},
\code{\link{occonet}},
\code{\link{occstatus}},
\code{\link{occtag}},
\code{\link{oth16}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{owngun}},
\code{\link{paattend}},
\code{\link{padeath}},
\code{\link{padeg}},
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{paeduc}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{palefull}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres10}},
\code{\link{papres105plus}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{parborn}},
\code{\link{parsol}},
\code{\link{partfull}},
\code{\link{partyid}},
\code{\link{pawrkgrw}},
\code{\link{pawrkslf}},
\code{\link{perla}},
\code{\link{phone}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{pistol}},
\code{\link{polabuse}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{polhitok}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{polmurdr}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{popespks}},
\code{\link{popular}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{premarsx}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres12}},
\code{\link{pres68}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{pres96}},
\code{\link{prestg10}},
\code{\link{prestg105plus}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{preteen}},
\code{\link{preteen_7222}},
\code{\link{preteen_exp}},
\code{\link{preteen_exp_22}},
\code{\link{privacy}},
\code{\link{prottemp}},
\code{\link{prowar}},
\code{\link{quitsmk}},
\code{\link{racavoid}},
\code{\link{raccare}},
\code{\link{racchng}},
\code{\link{racchurh}},
\code{\link{racclos}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racdif4}},
\code{\link{racdif5}},
\code{\link{racdif5x}},
\code{\link{racdin}},
\code{\link{racdis}},
\code{\link{race}},
\code{\link{raceacs1}},
\code{\link{raceacs10}},
\code{\link{raceacs11}},
\code{\link{raceacs12}},
\code{\link{raceacs13}},
\code{\link{raceacs14}},
\code{\link{raceacs14_2122}},
\code{\link{raceacs15}},
\code{\link{raceacs2}},
\code{\link{raceacs4}},
\code{\link{raceacs6}},
\code{\link{raceacs8}},
\code{\link{racfew}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{rachome}},
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racinfo}},
\code{\link{racinteg}},
\code{\link{racjob}},
\code{\link{racjoin}},
\code{\link{raclive}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racmix}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racobjct}},
\code{\link{racocc}},
\code{\link{racopen}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{racpres}},
\code{\link{racpush}},
\code{\link{racquit}},
\code{\link{racschol}},
\code{\link{racseg}},
\code{\link{racsubgv}},
\code{\link{racsubs}},
\code{\link{racsups}},
\code{\link{racteach}},
\code{\link{racthink}},
\code{\link{ractrust}},
\code{\link{racwrite}},
\code{\link{radiohrs}},
\code{\link{rank}()},
\code{\link{rank10}},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{reg16}},
\code{\link{reg16_7222}},
\code{\link{region}},
\code{\link{region_7222}},
\code{\link{relgrade}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{res16}},
\code{\link{resdefwk}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{restact}},
\code{\link{richwork}},
\code{\link{rifle}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{robbry}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{rowngun}},
\code{\link{rushed}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfin}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{satjob}},
\code{\link{school}},
\code{\link{sector}},
\code{\link{sex}},
\code{\link{sexeduc}},
\code{\link{sexnow2}},
\code{\link{shotgun}},
\code{\link{sibdeath}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{size}},
\code{\link{smoke}},
\code{\link{smokecig}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{spanking}},
\code{\link{spattend}},
\code{\link{spbarate}},
\code{\link{spdeath}},
\code{\link{spdeg}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spdipged}},
\code{\link{spdotdat}},
\code{\link{spdotged}},
\code{\link{spdotpeo}},
\code{\link{spdotpre}},
\code{\link{spdotsvp}},
\code{\link{spdotthn}},
\code{\link{spdue}},
\code{\link{speduc}},
\code{\link{speftotlt}},
\code{\link{spevwork}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spgradtounder}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spirits}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spkath}},
\code{\link{spkcom}},
\code{\link{spkhomo}},
\code{\link{spkmil}},
\code{\link{spkmslm}},
\code{\link{spkrac}},
\code{\link{spksoc}},
\code{\link{spnorelgsp}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sppaid}},
\code{\link{sppres}},
\code{\link{sppres10}},
\code{\link{sppres105plus}},
\code{\link{sppres80}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{spsector}},
\code{\link{spwrkslf}},
\code{\link{spwrksta}},
\code{\link{srcbelt}},
\code{\link{stresswk}},
\code{\link{strike}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{teenpill}},
\code{\link{teens}},
\code{\link{teens_7222}},
\code{\link{teens_exp}},
\code{\link{teens_exp_22}},
\code{\link{teensex}},
\code{\link{thnkself}},
\code{\link{ticket}},
\code{\link{tithing}},
\code{\link{transair}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{tvhours}},
\code{\link{typist}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrelat}},
\code{\link{unrelat_7222}},
\code{\link{upgrade}()},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}},
\code{\link{vetaid}},
\code{\link{vetfam}},
\code{\link{vetfamnw}},
\code{\link{visions}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}},
\code{\link{vote12}},
\code{\link{vote68}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}},
\code{\link{weekswrk}},
\code{\link{whencol}},
\code{\link{whenhs}},
\code{\link{widowed}},
\code{\link{wirtap}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}},
\code{\link{workhard}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{wrkwayup}},
\code{\link{xmarsex}},
\code{\link{xmovie}},
\code{\link{xmovie1}},
\code{\link{xnorcsiz}},
\code{\link{year}()},
\code{\link{youngen}}

Other Salience Of Issues: 
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}}

Other Science: 
\code{\link{accntsci}},
\code{\link{advfront}},
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{artseals}},
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{bettrlfe}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{biosci}},
\code{\link{bizbstgm}},
\code{\link{bizbstgw}},
\code{\link{bizinfgm}},
\code{\link{bizinfgw}},
\code{\link{boyorgrl}},
\code{\link{caremost}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{comorsci}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{eatgm}},
\code{\link{econsci}},
\code{\link{electron}},
\code{\link{engnrsci}},
\code{\link{erosion}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{extinct}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{gmbiz}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{goldfish}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{h2olife}},
\code{\link{harmgood}},
\code{\link{histsci}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{intspace}},
\code{\link{inuit}},
\code{\link{inuitway}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{leadsci}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{medagrgm}},
\code{\link{medagrsc}},
\code{\link{medbstgm}},
\code{\link{medbstsc}},
\code{\link{medinfgm}},
\code{\link{medinfsc}},
\code{\link{medsci}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{newsfrmy}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{noicecap}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{physcsci}},
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}},
\code{\link{polbstgm}},
\code{\link{polbstgw}},
\code{\link{polbstsc}},
\code{\link{polinfgm}},
\code{\link{polinfgw}},
\code{\link{polinfsc}},
\code{\link{radioact}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{salth2o}},
\code{\link{sciagrgw}},
\code{\link{scibnfts}},
\code{\link{scibstgw}},
\code{\link{scichng}},
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientda}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientsn}},
\code{\link{scifaith}},
\code{\link{scifrom}},
\code{\link{scifromy}},
\code{\link{scigrn}},
\code{\link{sciimp1}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{sciinfgo}},
\code{\link{sciinfgw}},
\code{\link{scimath}},
\code{\link{scimoral}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{scispec}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{sciworse}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}},
\code{\link{scresrch}},
\code{\link{sealevel}},
\code{\link{seeksci}},
\code{\link{seeksciy}},
\code{\link{seesand}},
\code{\link{socsci}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{toofast}},
\code{\link{trustsci}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}},
\code{\link{upbreath}},
\code{\link{viruses}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{weighing}},
\code{\link{whichsci}}
}
\concept{Core}
\concept{Salience Of Issues}
\concept{Science}
\keyword{variable}
