% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{samejob}
\alias{samejob}
\title{Current job same or different fromfirst job}
\source{
General Social Survey https://gss.norc.org
}
\description{
samejob
}
\details{
Question When you first came to work for your present employer (IF SELF-EMPLOYED: current organization), was it roughly in the job you have now or was it in a different job?
}
\section{Values}{

\itemize{
\item \code{1} job have now
\item \code{2} different job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1290/vshow}{GSS Data Explorer page for \code{samejob}}.
}

\seealso{
Other Work Organizations: 
\code{\link{jobyears}},
\code{\link{promotng}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}

Other Employment: 
\code{\link{jobyears}},
\code{\link{promotng}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}
}
\concept{Employment}
\concept{Work Organizations}
\keyword{variable}
