% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_143.R
\name{samevals}
\alias{samevals}
\title{My values and the org.'s values are very similar}
\source{
General Social Survey https://gss.norc.org
}
\description{
samevals
}
\details{
Question Please tell me how much you agree or disagree with the following statements. Would you say that you strongly agree, agree, disagree, or strongly disagree?
E. I find that my values and the organization's values are very similar.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1344/vshow}{GSS Data Explorer page for \code{samevals}}.
}

\seealso{
Other Work Organizations: 
\code{\link{chngeorg}},
\code{\link{helporg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{wrkmuch}}

Other Job: 
\code{\link{chngeorg}},
\code{\link{helporg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{stayorg1}},
\code{\link{stayorg2}}
}
\concept{Job}
\concept{Work Organizations}
\keyword{variable}
