% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{samework}
\alias{samework}
\title{Worked at other site with similar product}
\source{
General Social Survey https://gss.norc.org
}
\description{
samework
}
\details{
Question A. Please think about the (products made/services provided) at your work site. Have you worked at any other work site that (makes/provides) a similar kind of (products/services), i.e., similar by your perception?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2147/vshow}{GSS Data Explorer page for \code{samework}}.
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}

Other Job: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}
}
\concept{Job}
\concept{Jobs}
\keyword{variable}
