% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{sample}
\alias{sample}
\title{Sampling frame and method}
\source{
General Social Survey https://gss.norc.org
}
\description{
sample
}
\details{
Question The GSS has employed various sampling procedures and frames. Appendix A gives full detail.
}
\section{Values}{

\itemize{
\item \code{1} 1960 sample frame, block quota
\item \code{2} 1970 sample frame, block quota
\item \code{3} 1970 sample frame, full probability
\item \code{4} 1970 sample frame, full probability black oversample
\item \code{5} 1980 black sample frame, full probability black oversample
\item \code{6} 1980 sample frame, full probability
\item \code{7} 1980 sample frame, full probability black oversample
\item \code{8} 1990 sample frame, full probability
\item \code{9} 2000 sample frame, full probability
\item \code{10} 2010 sample frame, full probability
\item \code{11} 2020 address based sample, full probability
\item \code{12} 2022 amerispeak frame, full probability black/hispanic/aanhpi oversample
\item \code{13} 2020 sample frame, full probability
\item \code{14} 2024 amerispeak frame, full probability black/hispanic/aanhpi oversample
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{sampcode}},
\code{\link{version}}
}
\concept{Sampling and Weights}
\keyword{variable}
