% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_606.R
\name{saqissp}
\alias{saqissp}
\title{R did issp module}
\source{
General Social Survey https://gss.norc.org
}
\description{
saqissp
}
\details{
Question A. R did ISSP Module
}
\section{Values}{

\itemize{
\item \verb{[1]} r did issp
\item \verb{[2]} administered by interviewer
\item \verb{[3]} promised, but did not do
\item \verb{[4]} r did not do issp
\item \verb{[5]} r did, but not sure mode
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998, 2000, 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5556/vshow}{GSS Data Explorer page for \code{saqissp}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab r did issp \tab administered by interviewer \tab promised, but did not do \tab r did not do issp \tab r did, but not sure mode \tab Total \cr
   1996 \tab - \tab 1858 \tab 521 \tab 74 \tab 131 \tab 320 \tab 2904 \cr
   1998 \tab - \tab 1859 \tab 342 \tab 94 \tab 231 \tab 306 \tab 2832 \cr
   2000 \tab - \tab 1920 \tab 212 \tab 107 \tab 162 \tab 416 \tab 2817 \cr
   2002 \tab - \tab 1765 \tab 555 \tab - \tab 445 \tab - \tab 2765 \cr
   2004 \tab 1340 \tab 882 \tab 334 \tab - \tab 256 \tab - \tab 2812 \cr
   Total \tab 1340 \tab 8284 \tab 1964 \tab 275 \tab 1225 \tab 1042 \tab 14130 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{intid}},
\code{\link{saqgene}},
\code{\link{saqsex}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}
}
\concept{Paradata}
\keyword{variable}
