% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_192.R
\name{satcar}
\alias{satcar}
\title{How satisfied are you with your purchase?}
\source{
General Social Survey https://gss.norc.org
}
\description{
satcar
}
\details{
Question How satisfied are you with your purchase?
}
\section{Values}{

\itemize{
\item \verb{[1]} not too satisfied
\item \verb{[2]} pretty satisfied
\item \verb{[3]} extremely satisfied
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1811/vshow}{GSS Data Explorer page for \code{satcar}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab not too satisfied \tab pretty satisfied \tab extremely satisfied \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2497 \tab 36 \tab 157 \tab 203 \tab 1 \tab 10 \tab 2904 \cr
   Total \tab 2497 \tab 36 \tab 157 \tab 203 \tab 1 \tab 10 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{reldealr}}
}
\concept{Markets}
\keyword{variable}
