% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{satdemoc}
\alias{satdemoc}
\title{R satisfied with way democracy works in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
satdemoc
}
\details{
Question On the whole, are you very satisfied, fairly satisfied, not very satisfied, or not at all satisfied with the way democracy works in the United States?
}
\section{Values}{

\itemize{
\item \verb{[1]} very satisfied
\item \verb{[2]} fairly satisfied
\item \verb{[3]} not very satisfied
\item \verb{[4]} not at all satisfied
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2667/vshow}{GSS Data Explorer page for \code{satdemoc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very satisfied \tab fairly satisfied \tab not very satisfied \tab not at all satisfied \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 235 \tab 854 \tab 201 \tab 79 \tab 36 \tab 14 \tab 2817 \cr
   Total \tab 1398 \tab 235 \tab 854 \tab 201 \tab 79 \tab 36 \tab 14 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{freenow}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}}
}
\concept{ANES}
\concept{Freedom}
\keyword{variable}
