% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{satfam7}
\alias{satfam7}
\title{Family satisfaction in general}
\source{
General Social Survey https://gss.norc.org
}
\description{
satfam7
}
\details{
Question B. All things considered, how satisfied are you with your family life?
}
\section{Values}{

\itemize{
\item \code{1} completely satisfied
\item \code{2} very satisfied
\item \code{3} fairly satisfied
\item \code{4} neither satisfied nor dissatisfied
\item \code{5} fairly dissatisfied
\item \code{6} very dissatisfied
\item \code{7} completely dissatisfied
\item \code{8} does not apply
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012, 2018 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4453/vshow}{GSS Data Explorer page for \code{satfam7}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab completely satisfied \tab very satisfied \tab fairly satisfied \tab neither satisfied nor dissatisfied \tab fairly dissatisfied \tab very dissatisfied \tab completely dissatisfied \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1594 \tab 265 \tab 467 \tab 286 \tab 70 \tab 31 \tab 20 \tab 9 \tab 19 \tab 4 \tab 2765 \cr
   2012 \tab 672 \tab 254 \tab 548 \tab 316 \tab 72 \tab 41 \tab 22 \tab 10 \tab 13 \tab 26 \tab 1974 \cr
   2018 \tab 1173 \tab 331 \tab 450 \tab 254 \tab 60 \tab 40 \tab 23 \tab 9 \tab 3 \tab 5 \tab 2348 \cr
   Total \tab 3439 \tab 850 \tab 1465 \tab 856 \tab 202 \tab 112 \tab 65 \tab 28 \tab 35 \tab 35 \tab 7087 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
