% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{satfixes}
\alias{satfixes}
\title{How satisfied were you with the work}
\source{
General Social Survey https://gss.norc.org
}
\description{
satfixes
}
\details{
Question How satisfied were you with the work that was done for you?
}
\section{Values}{

\itemize{
\item \verb{[1]} not too satisfied
\item \verb{[2]} pretty satisfied
\item \verb{[3]} extremely satisfied
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1848/vshow}{GSS Data Explorer page for \code{satfixes}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab not too satisfied \tab pretty satisfied \tab extremely satisfied \tab no answer \tab Total \cr
   1996 \tab 2353 \tab 60 \tab 322 \tab 166 \tab 3 \tab 2904 \cr
   Total \tab 2353 \tab 60 \tab 322 \tab 166 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{fixer1}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{relfixer}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
