% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{sathobby}
\alias{sathobby}
\title{Non-working activities, hobbies}
\source{
General Social Survey https://gss.norc.org
}
\description{
sathobby
}
\details{
Question For each area of life I am going to name, tell me the number that shows how much satisfaction you get from that area. READ ITEMS A-E. CIRCLE ONE CODE FOR EACH.
B. Your non-working activities--hobbies and so on.
}
\section{Values}{

\itemize{
\item \code{1} very great deal
\item \code{2} great deal
\item \code{3} quite a bit
\item \code{4} a fair amount
\item \code{5} some
\item \code{6} a little
\item \code{7} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/444/vshow}{GSS Data Explorer page for \code{sathobby}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab very great deal \tab great deal \tab quite a bit \tab a fair amount \tab some \tab a little \tab none \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 383 \tab 430 \tab 251 \tab 192 \tab 78 \tab 87 \tab 66 \tab 11 \tab 6 \tab 1504 \cr
   1974 \tab - \tab 344 \tab 467 \tab 239 \tab 196 \tab 88 \tab 71 \tab 74 \tab 5 \tab - \tab 1484 \cr
   1975 \tab - \tab 359 \tab 482 \tab 240 \tab 182 \tab 83 \tab 70 \tab 61 \tab 9 \tab 4 \tab 1490 \cr
   1976 \tab - \tab 332 \tab 524 \tab 267 \tab 163 \tab 84 \tab 74 \tab 45 \tab 6 \tab 4 \tab 1499 \cr
   1977 \tab - \tab 386 \tab 447 \tab 292 \tab 206 \tab 75 \tab 66 \tab 47 \tab 9 \tab 2 \tab 1530 \cr
   1978 \tab - \tab 365 \tab 520 \tab 278 \tab 190 \tab 59 \tab 54 \tab 54 \tab 6 \tab 6 \tab 1532 \cr
   1980 \tab - \tab 399 \tab 500 \tab 230 \tab 161 \tab 72 \tab 60 \tab 38 \tab 7 \tab 1 \tab 1468 \cr
   1982 \tab - \tab 386 \tab 536 \tab 348 \tab 276 \tab 122 \tab 97 \tab 78 \tab 12 \tab 5 \tab 1860 \cr
   1983 \tab - \tab 326 \tab 507 \tab 319 \tab 219 \tab 103 \tab 52 \tab 57 \tab 13 \tab 3 \tab 1599 \cr
   1984 \tab - \tab 370 \tab 522 \tab 253 \tab 168 \tab 65 \tab 45 \tab 35 \tab 6 \tab 9 \tab 1473 \cr
   1986 \tab - \tab 268 \tab 497 \tab 274 \tab 191 \tab 76 \tab 71 \tab 71 \tab 9 \tab 13 \tab 1470 \cr
   1987 \tab - \tab 424 \tab 619 \tab 319 \tab 218 \tab 92 \tab 76 \tab 47 \tab 13 \tab 11 \tab 1819 \cr
   1988 \tab 484 \tab 213 \tab 359 \tab 181 \tab 117 \tab 56 \tab 35 \tab 30 \tab 4 \tab 2 \tab 1481 \cr
   1989 \tab 502 \tab 243 \tab 368 \tab 178 \tab 110 \tab 53 \tab 40 \tab 34 \tab 5 \tab 4 \tab 1537 \cr
   1990 \tab 473 \tab 205 \tab 317 \tab 154 \tab 105 \tab 52 \tab 27 \tab 32 \tab 5 \tab 2 \tab 1372 \cr
   1991 \tab 500 \tab 223 \tab 350 \tab 178 \tab 133 \tab 58 \tab 39 \tab 31 \tab 3 \tab 2 \tab 1517 \cr
   1993 \tab 549 \tab 252 \tab 345 \tab 193 \tab 125 \tab 63 \tab 40 \tab 28 \tab 4 \tab 7 \tab 1606 \cr
   1994 \tab 2481 \tab 115 \tab 175 \tab 93 \tab 58 \tab 33 \tab 14 \tab 21 \tab - \tab 2 \tab 2992 \cr
   Total \tab 4989 \tab 5593 \tab 7965 \tab 4287 \tab 3010 \tab 1312 \tab 1018 \tab 849 \tab 127 \tab 83 \tab 29233 \cr
}
}

\seealso{
Other Core: 
\code{\link{fair}},
\code{\link{hapcohab}},
\code{\link{health}},
\code{\link{helpful}},
\code{\link{life}},
\code{\link{satcity}},
\code{\link{trust}},
\code{\link{trusty}}
}
\concept{Core}
\keyword{variable}
