% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{satjob}
\alias{satjob}
\title{Work satisfaction}
\source{
General Social Survey https://gss.norc.org
}
\description{
satjob
}
\details{
Question On the whole, how satisfied are you with the work you do--would you say you are very satisfied, moderately satisfied,
a little dissatisfied, or very dissatisfied?
}
\section{Values}{

\itemize{
\item \code{1} very satisfied
\item \code{2} moderately satisfied
\item \code{3} a little dissatisfied
\item \code{4} very dissatisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/548/vshow}{GSS Data Explorer page for \code{satjob}}.

Counts by year:\tabular{llllllllll}{
   year \tab very satisfied \tab moderately satisfied \tab a little dissatisfied \tab very dissatisfied \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 458 \tab 347 \tab 106 \tab 33 \tab 3 \tab 660 \tab 6 \tab - \tab 1613 \cr
   1973 \tab 564 \tab 436 \tab 92 \tab 49 \tab 2 \tab 231 \tab 130 \tab - \tab 1504 \cr
   1974 \tab 586 \tab 458 \tab 122 \tab 57 \tab 5 \tab 224 \tab 32 \tab - \tab 1484 \cr
   1975 \tab 631 \tab 381 \tab 105 \tab 48 \tab 3 \tab 295 \tab 27 \tab - \tab 1490 \cr
   1976 \tab 617 \tab 407 \tab 108 \tab 53 \tab 3 \tab 294 \tab 17 \tab - \tab 1499 \cr
   1977 \tab 601 \tab 496 \tab 130 \tab 35 \tab 1 \tab 213 \tab 54 \tab - \tab 1530 \cr
   1978 \tab 651 \tab 463 \tab 108 \tab 58 \tab 1 \tab 222 \tab 29 \tab - \tab 1532 \cr
   1980 \tab 584 \tab 444 \tab 159 \tab 59 \tab 1 \tab 214 \tab 7 \tab - \tab 1468 \cr
   1982 \tab 686 \tab 562 \tab 160 \tab 84 \tab 3 \tab 301 \tab 64 \tab - \tab 1860 \cr
   1983 \tab 657 \tab 499 \tab 117 \tab 60 \tab - \tab 221 \tab 45 \tab - \tab 1599 \cr
   1984 \tab 555 \tab 423 \tab 146 \tab 84 \tab - \tab 208 \tab 57 \tab - \tab 1473 \cr
   1985 \tab 591 \tab 467 \tab 122 \tab 55 \tab 2 \tab 263 \tab 34 \tab - \tab 1534 \cr
   1986 \tab 570 \tab 458 \tab 104 \tab 30 \tab 1 \tab 257 \tab 50 \tab - \tab 1470 \cr
   1987 \tab 617 \tab 559 \tab 186 \tab 77 \tab 1 \tab 324 \tab 55 \tab - \tab 1819 \cr
   1988 \tab 537 \tab 455 \tab 117 \tab 44 \tab 2 \tab 290 \tab 36 \tab - \tab 1481 \cr
   1989 \tab 557 \tab 472 \tab 125 \tab 52 \tab 2 \tab 287 \tab 42 \tab - \tab 1537 \cr
   1990 \tab 481 \tab 413 \tab 109 \tab 38 \tab - \tab 263 \tab 68 \tab - \tab 1372 \cr
   1991 \tab 510 \tab 482 \tab 108 \tab 49 \tab 3 \tab 296 \tab 69 \tab - \tab 1517 \cr
   1993 \tab 532 \tab 505 \tab 133 \tab 58 \tab 1 \tab 329 \tab 48 \tab - \tab 1606 \cr
   1994 \tab 1078 \tab 932 \tab 247 \tab 81 \tab 7 \tab 551 \tab 96 \tab - \tab 2992 \cr
   1996 \tab 1041 \tab 918 \tab 253 \tab 101 \tab 3 \tab 488 \tab 100 \tab - \tab 2904 \cr
   1998 \tab 1067 \tab 851 \tab 223 \tab 75 \tab 7 \tab 527 \tab 82 \tab - \tab 2832 \cr
   2000 \tab 981 \tab 918 \tab 186 \tab 77 \tab 8 \tab 542 \tab 105 \tab - \tab 2817 \cr
   2002 \tab 534 \tab 386 \tab 100 \tab 41 \tab 13 \tab 1689 \tab 2 \tab - \tab 2765 \cr
   2004 \tab 702 \tab 507 \tab 125 \tab 63 \tab 16 \tab 1375 \tab 24 \tab - \tab 2812 \cr
   2006 \tab 1073 \tab 823 \tab 192 \tab 89 \tab 36 \tab 2237 \tab 60 \tab - \tab 4510 \cr
   2008 \tab 778 \tab 570 \tab 135 \tab 53 \tab 19 \tab 453 \tab 15 \tab - \tab 2023 \cr
   2010 \tab 746 \tab 565 \tab 155 \tab 62 \tab 20 \tab 477 \tab 19 \tab - \tab 2044 \cr
   2012 \tab 732 \tab 539 \tab 136 \tab 50 \tab 11 \tab 481 \tab 25 \tab - \tab 1974 \cr
   2014 \tab 956 \tab 686 \tab 161 \tab 92 \tab 12 \tab 626 \tab 5 \tab - \tab 2538 \cr
   2016 \tab 900 \tab 645 \tab 167 \tab 67 \tab 15 \tab 1064 \tab 9 \tab - \tab 2867 \cr
   2018 \tab 844 \tab 656 \tab 174 \tab 65 \tab 8 \tab 574 \tab 27 \tab - \tab 2348 \cr
   2021 \tab 1198 \tab 1119 \tab 305 \tab 112 \tab 5 \tab 1199 \tab 60 \tab 34 \tab 4032 \cr
   2022 \tab 1103 \tab 1012 \tab 241 \tab 107 \tab 7 \tab 1053 \tab 8 \tab 13 \tab 3544 \cr
   Total \tab 24718 \tab 19854 \tab 5157 \tab 2158 \tab 221 \tab 18728 \tab 1507 \tab 47 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{jobfind}},
\code{\link{jobinc}},
\code{\link{joblose}},
\code{\link{jobsec}},
\code{\link{richwork}}
}
\concept{Core}
\keyword{variable}
