% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{satjob7}
\alias{satjob7}
\title{Job satisfaction in general}
\source{
General Social Survey https://gss.norc.org
}
\description{
satjob7
}
\details{
Question A. All things considered, how satisfied are you with your (main) job?
}
\section{Values}{

\itemize{
\item \verb{[1]} completely satisfied
\item \verb{[2]} very satisfied
\item \verb{[3]} fairly satisfied
\item \verb{[4]} neither satisfied nor dissatisfied
\item \verb{[5]} fairly dissatisfied
\item \verb{[6]} very dissatisfied
\item \verb{[7]} completely dissatisfied
\item \verb{[8]} does not apply
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4452/vshow}{GSS Data Explorer page for \code{satjob7}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab completely satisfied \tab very satisfied \tab fairly satisfied \tab neither satisfied nor dissatisfied \tab fairly dissatisfied \tab very dissatisfied \tab completely dissatisfied \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1891 \tab 127 \tab 289 \tab 264 \tab 53 \tab 47 \tab 29 \tab 11 \tab 45 \tab 9 \tab 2765 \cr
   2012 \tab 1194 \tab 122 \tab 264 \tab 260 \tab 54 \tab 40 \tab 13 \tab 9 \tab 1 \tab 17 \tab 1974 \cr
   Total \tab 3085 \tab 249 \tab 553 \tab 524 \tab 107 \tab 87 \tab 42 \tab 20 \tab 46 \tab 26 \tab 4739 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
