% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{satsoc}
\alias{satsoc}
\title{R's satisfaction with social activities and relationships}
\source{
General Social Survey https://gss.norc.org
}
\description{
satsoc
}
\details{
Question In general, how would you rate your satisfaction with your social activities and relationships?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7166/vshow}{GSS Data Explorer page for \code{satsoc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 346 \tab 832 \tab 758 \tab 309 \tab 82 \tab 5 \tab 16 \tab - \tab 2348 \cr
   2021 \tab 380 \tab 362 \tab 1125 \tab 1283 \tab 645 \tab 207 \tab 1 \tab 1 \tab 28 \tab 4032 \cr
   Total \tab 380 \tab 708 \tab 1957 \tab 2041 \tab 954 \tab 289 \tab 6 \tab 17 \tab 28 \tab 6380 \cr
}
}

\keyword{variable}
