% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{savejobs}
\alias{savejobs}
\title{Govt support for declining industries}
\source{
General Social Survey https://gss.norc.org
}
\description{
savejobs
}
\details{
Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it.
G. Supporting declining industries to protect jobs.
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor of
\item \code{2} in favor
\item \code{3} neither in favor nor against
\item \code{4} against
\item \code{5} strongly against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4055/vshow}{GSS Data Explorer page for \code{savejobs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly in favor of \tab in favor \tab neither in favor nor against \tab against \tab strongly against \tab no answer \tab don't know \tab Total \cr
   1985 \tab 857 \tab 118 \tab 225 \tab 147 \tab 144 \tab 34 \tab 9 \tab - \tab 1534 \cr
   1990 \tab 155 \tab 177 \tab 436 \tab 341 \tab 188 \tab 44 \tab 30 \tab 1 \tab 1372 \cr
   1996 \tab 1572 \tab 178 \tab 457 \tab 326 \tab 263 \tab 66 \tab 38 \tab 4 \tab 2904 \cr
   2006 \tab 2992 \tab 369 \tab 651 \tab 232 \tab 203 \tab 45 \tab 2 \tab 16 \tab 4510 \cr
   2016 \tab 1477 \tab 252 \tab 595 \tab 292 \tab 189 \tab 40 \tab 9 \tab 13 \tab 2867 \cr
   Total \tab 7053 \tab 1094 \tab 2364 \tab 1338 \tab 987 \tab 229 \tab 88 \tab 34 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{poleff10}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{spenviro}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
