% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_168.R
\name{sbalive1}
\alias{sbalive1}
\title{R 1st sibling alive?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbalive1
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
E1. Is first sibling alive?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1572/vshow}{GSS Data Explorer page for \code{sbalive1}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbrel9}}

Other Siblings: 
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbrel9}}
}
\concept{Family Mobility}
\concept{Siblings}
\keyword{variable}
