% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_168.R
\name{sbalive6}
\alias{sbalive6}
\title{R 6th sibling alive?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbalive6
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
E6. Is sixth sibling alive?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1577/vshow}{GSS Data Explorer page for \code{sbalive6}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2398 \tab 465 \tab 95 \tab 4 \tab 30 \tab 2992 \cr
   Total \tab 2398 \tab 465 \tab 95 \tab 4 \tab 30 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbrel9}}
}
\concept{Family Mobility}
\keyword{variable}
