% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_168.R
\name{sbalive8}
\alias{sbalive8}
\title{R 8th sibling alive?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbalive8
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
E8. Is eighth sibling alive?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1579/vshow}{GSS Data Explorer page for \code{sbalive8}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2702 \tab 216 \tab 48 \tab 2 \tab 24 \tab 2992 \cr
   Total \tab 2702 \tab 216 \tab 48 \tab 2 \tab 24 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive9}},
\code{\link{sbrel9}}
}
\concept{Family Mobility}
\keyword{variable}
