% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbeduc}
\alias{sbeduc}
\title{Highest schl grade of selected r sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbeduc
}
\details{
Question SIBLING'S EDUCATION
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1582/vshow}{GSS Data Explorer page for \code{sbeduc}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab Total \cr
   1994 \tab 378 \tab 8 \tab 118 \tab 130 \tab 993 \tab 134 \tab 253 \tab 74 \tab 409 \tab 39 \tab 88 \tab 20 \tab 56 \tab 4 \tab 9 \tab 6 \tab 20 \tab 22 \tab 75 \tab 45 \tab 111 \tab 2992 \cr
   Total \tab 378 \tab 8 \tab 118 \tab 130 \tab 993 \tab 134 \tab 253 \tab 74 \tab 409 \tab 39 \tab 88 \tab 20 \tab 56 \tab 4 \tab 9 \tab 6 \tab 20 \tab 22 \tab 75 \tab 45 \tab 111 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbdeg}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
