% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbevwork}
\alias{sbevwork}
\title{Did r sibling work for 1 or more yr?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbevwork
}
\details{
Question Has (s/he) ever worked at one job for as long as one year?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1586/vshow}{GSS Data Explorer page for \code{sbevwork}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
