% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbind80}
\alias{sbind80}
\title{R sibling industry code (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbind80
}
\details{
Question Respondent's Sibling's Occupation, Prestige, and Industry
A. What kind of work (is/was) (BROTHER'S NAME/SISTER'S NAME) doing? That is, what was (his/her) job called?
OCCUPATION (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What (does/did) (s/he) actually do in that job? Tell me some of what (are/were) some of (his/her) main duties?
C. What kind of place did (s/he) work for?
INDUSTRY (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What (do/did) they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  (Is/Was) (he/she) self-employed or (does/did) (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[10]} agricultural production, crops
\item \verb{[11]} agricultural production, livestock
\item \verb{[20]} agricultural services, except horticultural
\item \verb{[21]} horticultural services
\item \verb{[30]} forestry
\item \verb{[31]} fishing, hunting, and trapping, etc.
\item \verb{[40]} metal mining
\item \verb{[41]} coal mining
\item \verb{[42]} crude petroleum and natural gas extraction
\item \verb{[50]} nonmetallic mining and quarrying, except fuel
\item \verb{[60]} construction
\item \verb{[100]} meat products
\item \verb{[101]} dairy products
\item \verb{[102]} canned and preserved fruits and vegetables
\item \verb{[110]} grain mill products
\item \verb{[111]} bakery products
\item \verb{[112]} sugar and confectionery products
\item \verb{[120]} beverage industries
\item \verb{[121]} miscellaneous food preparations and kindred products
\item \verb{[122]} not specified food industries
\item \verb{[130]} tobacco manufactures
\item \verb{[132]} knitting mills
\item \verb{[140]} dyeing and finishing textiles, except wool and knit goods
\item \verb{[141]} floor coverings, except hard surface
\item \verb{[142]} yarn, thread, and fabric mills
\item \verb{[150]} miscellaneous textile mill products
\item \verb{[151]} apparel and accessories, except knit
\item \verb{[152]} miscellaneous fabricated textile products
\item \verb{[160]} pulp, paper, and paperboard mills
\item \verb{[161]} miscellaneous paper and pulp products
\item \verb{[162]} paperboard containers and boxes
\item \verb{[171]} newspaper publishing and printing
\item \verb{[172]} printing, publishing, and allied industries, except newspapers
\item \verb{[180]} plastics, synthetics, and resins
\item \verb{[181]} drugs
\item \verb{[182]} soaps and cosmetics
\item \verb{[185]} paints, varnishes, and related products
\item \verb{[190]} agricultural chemicals
\item \verb{[191]} industrial and miscellaneous chemicals
\item \verb{[192]} petroleum and coal products
\item \verb{[200]} petroleum refining
\item \verb{[201]} miscellaneous petroleum and coal products
\item \verb{[210]} tires and inner tubes
\item \verb{[211]} other rubber products, and plastics footwear and belting
\item \verb{[212]} miscellaneous plastics products
\item \verb{[220]} leather tanning and finishing
\item \verb{[221]} footwear, except rubber and plastic
\item \verb{[222]} leather products, except footwear
\item \verb{[230]} logging
\item \verb{[231]} sawmills, planning mills, and millwork
\item \verb{[232]} wood buildings and mobile homes
\item \verb{[241]} miscellaneous wood products
\item \verb{[242]} furniture and fixtures
\item \verb{[250]} glass and glass products
\item \verb{[251]} cement, concrete, gypsum, and plaster products
\item \verb{[252]} structural clay products
\item \verb{[261]} pottery and related products
\item \verb{[262]} miscellaneous nonmetallic mineral and stone products
\item \verb{[270]} blast furnaces, steelworks, rolling and finishing mills
\item \verb{[271]} iron and steel foundries
\item \verb{[272]} primary aluminum industries
\item \verb{[280]} other primary metal industries
\item \verb{[281]} cutlery, hand tools, and other hardware
\item \verb{[282]} fabricated structural metal products
\item \verb{[290]} screw machine products
\item \verb{[291]} metal forgings and stampings
\item \verb{[292]} ordnance
\item \verb{[300]} miscellaneous fabricated metal products
\item \verb{[301]} not specified metal industries
\item \verb{[310]} engines and turbines
\item \verb{[311]} farm machinery and equipment
\item \verb{[312]} construction and material handling machines
\item \verb{[320]} metalworking machinery
\item \verb{[321]} office and accounting machines
\item \verb{[322]} electronic computing equipment
\item \verb{[331]} machinery, except electrical, n.e.c.
\item \verb{[332]} not specified machinery
\item \verb{[340]} household appliances
\item \verb{[341]} radio, tv, and communication equipment
\item \verb{[342]} electrical machinery, equipment, and supplies, n.e.c.
\item \verb{[350]} not specified electrical machinery, equipment, and supplies
\item \verb{[351]} motor vehicles and motor vehicle equipment
\item \verb{[352]} aircraft and parts
\item \verb{[360]} ship and boat building and repairing
\item \verb{[361]} railroad locomotives and equipment
\item \verb{[362]} guided missiles, space vehicles, and parts
\item \verb{[370]} cycles and miscellaneous transportation equipment
\item \verb{[371]} scientific and controlling instruments
\item \verb{[372]} optical and health services supplies
\item \verb{[380]} photographic equipment and supplies
\item \verb{[381]} watches, clocks, and clockwork operated devices
\item \verb{[382]} not specified professional equipment
\item \verb{[390]} toys, amusement, and sporting goods
\item \verb{[391]} miscellaneous manufacturing industries
\item \verb{[392]} not specified manufacturing industries
\item \verb{[400]} railroads
\item \verb{[401]} bus service and urban transit
\item \verb{[402]} taxicab service
\item \verb{[410]} trucking service
\item \verb{[411]} warehousing and storage
\item \verb{[412]} u.s. postal service
\item \verb{[420]} water transportation
\item \verb{[421]} air transportation
\item \verb{[422]} pipe lines, except natural gas
\item \verb{[432]} services incidental to transportation
\item \verb{[440]} radio and television broadcasting
\item \verb{[441]} telephone (wire and radio)
\item \verb{[442]} telegraph and miscellaneous communication service
\item \verb{[460]} electric light and power
\item \verb{[461]} gas and steam supply systems
\item \verb{[462]} electric and gas, and other combinations
\item \verb{[470]} water supply and irrigation
\item \verb{[471]} sanitary services
\item \verb{[472]} not specified utilities
\item \verb{[500]} motor vehicles and equipment
\item \verb{[501]} furniture and home furnishings
\item \verb{[502]} lumber and construction materials
\item \verb{[510]} sporting goods, toys, and hobby goods
\item \verb{[511]} metals and minerals, except petroleum
\item \verb{[512]} electrical goods
\item \verb{[521]} hardware, plumbing and heating supplies
\item \verb{[522]} not specified electrical and hardware products
\item \verb{[530]} machinery, equipment, and supplies
\item \verb{[531]} scrap and waste materials
\item \verb{[532]} miscellaneous wholesale, durable goods
\item \verb{[540]} paper and paper products
\item \verb{[541]} drugs, chemicals, and allied products
\item \verb{[542]} apparel, fabrics, and notions
\item \verb{[550]} groceries and related products
\item \verb{[551]} farm products-raw materials
\item \verb{[552]} petroleum products
\item \verb{[560]} alcoholic beverages
\item \verb{[561]} farm supplies
\item \verb{[562]} miscellaneous wholesale, nondurable goods
\item \verb{[571]} not specified wholesale trade
\item \verb{[580]} lumber and building material retailing
\item \verb{[581]} hardware stores
\item \verb{[582]} retail nurseries and garden stores
\item \verb{[590]} mobile home dealers
\item \verb{[591]} department stores
\item \verb{[592]} variety stores
\item \verb{[600]} miscellaneous general merchandise stores
\item \verb{[601]} grocery stores
\item \verb{[602]} dairy products stores
\item \verb{[610]} retail bakeries
\item \verb{[611]} food stores, n.e.c.
\item \verb{[612]} motor vehicle dealers
\item \verb{[620]} auto and home supply stores
\item \verb{[621]} gasoline service stations
\item \verb{[622]} miscellaneous vehicle dealers
\item \verb{[630]} apparel and accessory stores, except shoe
\item \verb{[631]} shoe stores
\item \verb{[632]} furniture and home furnishings stores
\item \verb{[640]} household appliances, tv, and radio stores
\item \verb{[641]} eating and drinking places
\item \verb{[642]} drug stores
\item \verb{[650]} liquor stores
\item \verb{[651]} sporting goods, bicycles, and hobby stores
\item \verb{[652]} book and stationery stores
\item \verb{[660]} jewelry stores
\item \verb{[661]} sewing, needlework, and piece goods stores
\item \verb{[662]} mail order houses
\item \verb{[670]} vending machine operators
\item \verb{[671]} direct selling establishments
\item \verb{[672]} fuel and ice dealers
\item \verb{[681]} retail florists
\item \verb{[682]} miscellaneous retail stores
\item \verb{[691]} not specified retail trade
\item \verb{[700]} banking
\item \verb{[701]} savings and loan associations
\item \verb{[702]} credit agencies, n.e.c.
\item \verb{[710]} security, commodity brokerage, and investment companies
\item \verb{[711]} insurance
\item \verb{[712]} real estate, including real estate-insurance-law offices
\item \verb{[721]} advertising
\item \verb{[722]} services to dwellings and other buildings
\item \verb{[730]} commercial research, development, and testing labs
\item \verb{[731]} personnel supply services
\item \verb{[732]} business management and consulting services
\item \verb{[740]} computer and data processing services
\item \verb{[741]} detective and protective services
\item \verb{[742]} business services, n.e.c.
\item \verb{[750]} automotive services, except repair
\item \verb{[751]} automotive repair shops
\item \verb{[752]} electrical repair shops
\item \verb{[760]} miscellaneous repair services
\item \verb{[761]} private households
\item \verb{[762]} hotels and motels
\item \verb{[770]} lodging places, except hotels and motels
\item \verb{[771]} laundry, cleaning, and garment services
\item \verb{[772]} beauty shops
\item \verb{[780]} barber shops
\item \verb{[781]} funeral service and crematories
\item \verb{[782]} shoe repair shops
\item \verb{[790]} dressmaking shops
\item \verb{[791]} miscellaneous personal services
\item \verb{[800]} theaters and motion pictures
\item \verb{[801]} bowling alleys, billiard and pool parlors
\item \verb{[802]} miscellaneous entertainment and recreation services
\item \verb{[812]} offices of physicians
\item \verb{[820]} offices of dentists
\item \verb{[821]} offices of chiropractors
\item \verb{[822]} offices of optometrists
\item \verb{[830]} offices of health practitioners, n.e.c.
\item \verb{[831]} hospitals
\item \verb{[832]} nursing and personal care facilities
\item \verb{[840]} health services, n.e.c.
\item \verb{[841]} legal services
\item \verb{[842]} elementary and secondary schools
\item \verb{[850]} colleges and universities
\item \verb{[851]} business, trade, and vocational schools
\item \verb{[852]} libraries
\item \verb{[860]} educational services, n.e.c.
\item \verb{[861]} job training and vocational rehabilitation services
\item \verb{[862]} child day care services
\item \verb{[870]} residential care facilities, without nursing
\item \verb{[871]} social services, n.e.c.
\item \verb{[872]} museums, art galleries and zoos
\item \verb{[880]} religious organizations
\item \verb{[881]} membership organizations
\item \verb{[882]} engineering, architectural, and surveying services
\item \verb{[890]} accounting, auditing, and bookkeeping services
\item \verb{[891]} noncommercial educational and scientific research
\item \verb{[892]} miscellaneous professional and related services
\item \verb{[900]} executive and legislative offices
\item \verb{[901]} general government, n.e.c.
\item \verb{[910]} justice, public order, and safety
\item \verb{[921]} public finance, taxation, and monetary policy
\item \verb{[922]} administration of human resources programs
\item \verb{[930]} administration of environmental quality and household programs
\item \verb{[931]} administration of economic programs
\item \verb{[932]} national security and international affairs
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{fstspyr}},
\code{\link{indfirst}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpicked}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{malive1}},
\code{\link{marnum}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{palive1}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind10}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sbocc80}},
\code{\link{sboccindv}},
\code{\link{sboccstatus}},
\code{\link{sbocctag}},
\code{\link{sbpicked}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{spfam16}},
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre}},
\code{\link{sphedpre10}},
\code{\link{sphedpre105plus}},
\code{\link{sphedslf}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spsibs}},
\code{\link{spyrborn}},
\code{\link{wrkslfst}}

Other Industry: 
\code{\link{fambiz}},
\code{\link{indfirst}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{kdind80}},
\code{\link{maind80}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{sphedind}},
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{wrkgovt}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wkfambiz}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}

Other Siblings: 
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sibnum}},
\code{\link{siborder}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{spsibs}}
}
\concept{Family Mobility}
\concept{Industry}
\concept{Occupation}
\concept{Siblings}
\keyword{variable}
