% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_170.R
\name{sboccstatus}
\alias{sboccstatus}
\title{Sbocc10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
sboccstatus
}
\details{
Question SBOCC10 Coding Status
}
\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{8} case not collected in 1994
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6021/vshow}{GSS Data Explorer page for \code{sboccstatus}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sbind10}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sboccindv}},
\code{\link{sbocctag}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}}
}
\concept{Family Mobility}
\keyword{variable}
