% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_167.R
\name{sbrel5}
\alias{sbrel5}
\title{R relation to r 5th sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbrel5
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
D5. Relation to R. of fifth sibling
}
\section{Values}{

\itemize{
\item \verb{[1]} full
\item \verb{[2]} adopted
\item \verb{[3]} step
\item \verb{[4]} half
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1567/vshow}{GSS Data Explorer page for \code{sbrel5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab full \tab adopted \tab step \tab half \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2158 \tab 644 \tab 5 \tab 47 \tab 93 \tab 2 \tab 43 \tab 2992 \cr
   Total \tab 2158 \tab 644 \tab 5 \tab 47 \tab 93 \tab 2 \tab 43 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}}
}
\concept{Family Mobility}
\keyword{variable}
