% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_168.R
\name{sbrel9}
\alias{sbrel9}
\title{R relation to r 9th sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbrel9
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
D9. Relation to R. of ninth sibling
}
\section{Values}{

\itemize{
\item \verb{[1]} full
\item \verb{[2]} adopted
\item \verb{[3]} step
\item \verb{[4]} half
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1571/vshow}{GSS Data Explorer page for \code{sbrel9}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab full \tab adopted \tab step \tab half \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2795 \tab 138 \tab 1 \tab 11 \tab 27 \tab 1 \tab 19 \tab 2992 \cr
   Total \tab 2795 \tab 138 \tab 1 \tab 11 \tab 27 \tab 1 \tab 19 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}}
}
\concept{Family Mobility}
\keyword{variable}
