% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_165.R
\name{sbsex3}
\alias{sbsex3}
\title{Sex of r 3rd sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbsex3
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
B3. Sex of third sibling.
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1547/vshow}{GSS Data Explorer page for \code{sbsex3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab male \tab female \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1297 \tab 802 \tab 852 \tab 1 \tab 40 \tab 2992 \cr
   Total \tab 1297 \tab 802 \tab 852 \tab 1 \tab 40 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}}
}
\concept{Family Mobility}
\keyword{variable}
