% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_166.R
\name{sbsex7}
\alias{sbsex7}
\title{Sex of r 7th sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbsex7
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
B7. Sex of seventh sibling.
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1551/vshow}{GSS Data Explorer page for \code{sbsex7}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab male \tab female \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2583 \tab 196 \tab 185 \tab 4 \tab 24 \tab 2992 \cr
   Total \tab 2583 \tab 196 \tab 185 \tab 4 \tab 24 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}}
}
\concept{Family Mobility}
\keyword{variable}
