% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_166.R
\name{sbsex8}
\alias{sbsex8}
\title{Sex of r 8th sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbsex8
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
B8. Sex of eighth sibling.
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1552/vshow}{GSS Data Explorer page for \code{sbsex8}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sbsex7}},
\code{\link{sbsex9}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}}

Other Sex: 
\code{\link{sbsex7}},
\code{\link{sbsex9}}

Other Siblings: 
\code{\link{sbsex7}},
\code{\link{sbsex9}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}}
}
\concept{Family Mobility}
\concept{Sex}
\concept{Siblings}
\keyword{variable}
