% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_166.R
\name{sbyrbrn2}
\alias{sbyrbrn2}
\title{Birth year of r 2nd sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbyrbrn2
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
C2. Birth year of second sibling.
}
\section{Values}{

\itemize{
\item \verb{[9996]} before 1976
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1555/vshow}{GSS Data Explorer page for \code{sbyrbrn2}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1880 \tab 1889 \tab 1894 \tab 1897 \tab 1898 \tab 1899 \tab 1900 \tab 1901 \tab 1902 \tab 1903 \tab 1904 \tab 1905 \tab 1906 \tab 1907 \tab 1908 \tab 1909 \tab 1910 \tab 1911 \tab 1912 \tab 1913 \tab 1914 \tab 1915 \tab 1916 \tab 1917 \tab 1918 \tab 1919 \tab 1920 \tab 1921 \tab 1922 \tab 1923 \tab 1924 \tab 1925 \tab 1926 \tab 1927 \tab 1928 \tab 1929 \tab 1930 \tab 1931 \tab 1932 \tab 1933 \tab 1934 \tab 1935 \tab 1936 \tab 1937 \tab 1938 \tab 1939 \tab 1940 \tab 1941 \tab 1942 \tab 1943 \tab 1944 \tab 1945 \tab 1946 \tab 1947 \tab 1948 \tab 1949 \tab 1950 \tab 1951 \tab 1952 \tab 1953 \tab 1954 \tab 1955 \tab 1956 \tab 1957 \tab 1958 \tab 1959 \tab 1960 \tab 1961 \tab 1962 \tab 1963 \tab 1964 \tab 1965 \tab 1966 \tab 1967 \tab 1968 \tab 1969 \tab 1970 \tab 1971 \tab 1972 \tab 1973 \tab 1974 \tab 1975 \tab 1976 \tab 1977 \tab 1978 \tab 1979 \tab 1980 \tab 1981 \tab 1982 \tab 1983 \tab 1984 \tab 1985 \tab 1986 \tab 1988 \tab 1989 \tab 1990 \tab before 1976 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 721 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab 3 \tab 5 \tab 2 \tab 6 \tab 6 \tab 4 \tab 6 \tab 7 \tab 5 \tab 12 \tab 15 \tab 11 \tab 12 \tab 10 \tab 23 \tab 15 \tab 10 \tab 23 \tab 16 \tab 16 \tab 27 \tab 18 \tab 24 \tab 20 \tab 15 \tab 15 \tab 13 \tab 15 \tab 22 \tab 17 \tab 11 \tab 24 \tab 23 \tab 14 \tab 15 \tab 21 \tab 17 \tab 26 \tab 30 \tab 22 \tab 39 \tab 32 \tab 35 \tab 32 \tab 28 \tab 48 \tab 37 \tab 49 \tab 38 \tab 52 \tab 57 \tab 59 \tab 40 \tab 70 \tab 57 \tab 60 \tab 53 \tab 46 \tab 39 \tab 52 \tab 39 \tab 53 \tab 50 \tab 47 \tab 34 \tab 45 \tab 31 \tab 23 \tab 31 \tab 31 \tab 13 \tab 22 \tab 22 \tab 18 \tab 7 \tab 14 \tab 12 \tab 11 \tab 10 \tab 9 \tab 6 \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 113 \tab 82 \tab 2992 \cr
   Total \tab 721 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab 3 \tab 5 \tab 2 \tab 6 \tab 6 \tab 4 \tab 6 \tab 7 \tab 5 \tab 12 \tab 15 \tab 11 \tab 12 \tab 10 \tab 23 \tab 15 \tab 10 \tab 23 \tab 16 \tab 16 \tab 27 \tab 18 \tab 24 \tab 20 \tab 15 \tab 15 \tab 13 \tab 15 \tab 22 \tab 17 \tab 11 \tab 24 \tab 23 \tab 14 \tab 15 \tab 21 \tab 17 \tab 26 \tab 30 \tab 22 \tab 39 \tab 32 \tab 35 \tab 32 \tab 28 \tab 48 \tab 37 \tab 49 \tab 38 \tab 52 \tab 57 \tab 59 \tab 40 \tab 70 \tab 57 \tab 60 \tab 53 \tab 46 \tab 39 \tab 52 \tab 39 \tab 53 \tab 50 \tab 47 \tab 34 \tab 45 \tab 31 \tab 23 \tab 31 \tab 31 \tab 13 \tab 22 \tab 22 \tab 18 \tab 7 \tab 14 \tab 12 \tab 11 \tab 10 \tab 9 \tab 6 \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 113 \tab 82 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}}
}
\concept{Family Mobility}
\keyword{variable}
