% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_167.R
\name{sbyrbrn8}
\alias{sbyrbrn8}
\title{Birth year of r 8th sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbyrbrn8
}
\details{
Question Now I have some questions about your sisters and brothers. These include your full, adopted, step, or half siblings. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST SIBLINGS IN TABLE TO RIGHT).
C8. Birth year of eighth sibling.
}
\section{Values}{

\itemize{
\item \code{9996} before 1976
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1561/vshow}{GSS Data Explorer page for \code{sbyrbrn8}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbyrbrn9}}

Other Date Of Birth: 
\code{\link{sbyrbrn9}}

Other Siblings: 
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbyrbrn9}}
}
\concept{Date Of Birth}
\concept{Family Mobility}
\concept{Siblings}
\keyword{variable}
