% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{schlhead}
\alias{schlhead}
\title{Did r correctly name local school head}
\source{
General Social Survey https://gss.norc.org
}
\description{
schlhead
}
\details{
Question What is the name of the head of the local school system? (CORRECT ANSWERS INCLUDE THE NAME OF THE HEAD OF THE LOCAL SCHOOL BOARD OR THE NAME OF THE SUPERINTENDENT OF SCHOOLS).
}
\section{Values}{

\itemize{
\item \verb{[1]} correct
\item \verb{[2]} incorrect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1037/vshow}{GSS Data Explorer page for \code{schlhead}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab correct \tab incorrect \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 557 \tab 66 \tab 1162 \tab 34 \tab 1819 \cr
   Total \tab 0 \tab 557 \tab 66 \tab 1162 \tab 34 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
