% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{schlpray}
\alias{schlpray}
\title{Should there be daily prayer in public schools?}
\source{
General Social Survey https://gss.norc.org
}
\description{
schlpray
}
\details{
Question In your opinion, should there be daily prayers in all public schools?
}
\section{Values}{

\itemize{
\item \code{1} yes, definitely
\item \code{2} yes, probably
\item \code{3} no, probably
\item \code{4} no, definitely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4680/vshow}{GSS Data Explorer page for \code{schlpray}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{feelrel}},
\code{\link{prayfreq}},
\code{\link{relactiv}},
\code{\link{relactiv1}}

Other Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{feelrel}},
\code{\link{prayfreq}},
\code{\link{relactiv}},
\code{\link{relactiv1}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
