% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_682.R
\name{schshoot}
\alias{schshoot}
\title{Would x carry out a school shooting}
\source{
General Social Survey https://gss.norc.org
}
\description{
schshoot
}
\details{
Question How likely in your opinion is it (VIGNAME) would carry out a school shooting?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8141/vshow}{GSS Data Explorer page for \code{schshoot}}.
}

\seealso{
Other Mental Health: 
\code{\link{cutself}},
\code{\link{fighpeer}},
\code{\link{foodalc}},
\code{\link{negself}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{takelife}}
}
\concept{Mental Health}
\keyword{variable}
