% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_266.R
\name{sci30}
\alias{sci30}
\title{R use science site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
sci30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
S. Science?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} 1-2 times
\item \verb{[3]} 3-5 times
\item \verb{[4]} more than 5 times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2412/vshow}{GSS Data Explorer page for \code{sci30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 376 \tab 147 \tab 79 \tab 65 \tab 75 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 618 \tab 200 \tab 84 \tab 86 \tab 94 \tab 2 \tab 2765 \cr
   2004 \tab 2122 \tab 440 \tab 115 \tab 68 \tab 54 \tab 12 \tab 1 \tab 2812 \cr
   Total \tab 5878 \tab 1434 \tab 462 \tab 231 \tab 205 \tab 181 \tab 3 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{invest12}},
\code{\link{person30}},
\code{\link{porn30}},
\code{\link{work12}}
}
\concept{Information Society}
\keyword{variable}
