% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{scibnftsv}
\alias{scibnftsv}
\title{Benefits of sci research outweigh harmful results (with volunteered response on}
\source{
General Social Survey https://gss.norc.org
}
\description{
scibnftsv
}
\details{
Question Now for another type of question. People have frequently noted that scientific research has produced benefits and harmful results. Would you say that, on balance, the benefits of scientific research have outweighed the harmful results, or have the harmful results of scientific research been greater than its benefits?
}
\section{Values}{

\itemize{
\item \code{1} benefits greater
\item \code{2} about equal (phone mode only: volunteered)
\item \code{3} harmful results greater
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7603/vshow}{GSS Data Explorer page for \code{scibnftsv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab benefits greater \tab about equal (phone mode only: volunteered) \tab harmful results greater \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2159 \tab 1078 \tab 698 \tab 62 \tab 12 \tab 1 \tab 22 \tab 4032 \cr
   Total \tab 2159 \tab 1078 \tab 698 \tab 62 \tab 12 \tab 1 \tab 22 \tab 4032 \cr
}
}

\keyword{variable}
