% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{scibstgw}
\alias{scibstgw}
\title{Best policy vs narrow interest: environmental scientist}
\source{
General Social Survey https://gss.norc.org
}
\description{
scibstgw
}
\details{
Question Please look at Card B5. When making policy recommendations about global warming, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole versus what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
A. Environmental scientists
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3362/vshow}{GSS Data Explorer page for \code{scibstgw}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab what is best for the country \tab 2 \tab 3 \tab 4 \tab own narrow interests \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 368 \tab 245 \tab 159 \tab 50 \tab 61 \tab 43 \tab 2 \tab 4510 \cr
   2010 \tab 1569 \tab 193 \tab 108 \tab 84 \tab 34 \tab 29 \tab 23 \tab 4 \tab 2044 \cr
   Total \tab 5151 \tab 561 \tab 353 \tab 243 \tab 84 \tab 90 \tab 66 \tab 6 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgw}},
\code{\link{bizinfgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
