% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_387.R
\name{scientdo}
\alias{scientdo}
\title{Know what scientists do}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientdo
}
\details{
Question D. Next, I’d like to ask how you rate your knowledge of what scientists do day-to-day in their jobs. would you say your knowledge of what scientists do day-to-day on their jobs is Excellent, Good, Fair, Poor, Very Poor?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} good
\item \code{3} fair
\item \code{4} poor
\item \code{5} very poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3510/vshow}{GSS Data Explorer page for \code{scientdo}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientdn}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientre}}

Other Science: 
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientdn}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientr}}
}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
