% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_387.R
\name{scientfu}
\alias{scientfu}
\title{Scientists don? Have fun}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientfu
}
\details{
Question D. Scientists don't get as much fun out of life as other people Do. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3514/vshow}{GSS Data Explorer page for \code{scientfu}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1457 \tab 14 \tab 101 \tab 300 \tab 47 \tab 52 \tab 3 \tab 1974 \cr
   Total \tab 1457 \tab 14 \tab 101 \tab 300 \tab 47 \tab 52 \tab 3 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientre}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
