% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_387.R
\name{scienthe}
\alias{scienthe}
\title{Scientists help solve prob}
\source{
General Social Survey https://gss.norc.org
}
\description{
scienthe
}
\details{
Question E. Scientists are helping to solve challenging problems. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2022 \cr
   B/C/- \tab 2012, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3515/vshow}{GSS Data Explorer page for \code{scienthe}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 1457 \tab 116 \tab 377 \tab 2 \tab 1 \tab 18 \tab 3 \tab - \tab 1974 \cr
   2016 \tab 1477 \tab 371 \tab 930 \tab 52 \tab 9 \tab 19 \tab 9 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 324 \tab 767 \tab 66 \tab 5 \tab 12 \tab 1 \tab - \tab 2348 \cr
   2021 \tab 2155 \tab 653 \tab 1092 \tab 97 \tab 13 \tab 3 \tab 1 \tab 18 \tab 4032 \cr
   2022 \tab 2309 \tab 428 \tab 738 \tab 54 \tab 9 \tab - \tab - \tab 6 \tab 3544 \cr
   Total \tab 8571 \tab 1892 \tab 3904 \tab 271 \tab 37 \tab 52 \tab 14 \tab 24 \tab 14765 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientre}}

Other GSS Next: 
\code{\link{scientbe}},
\code{\link{scientgo}}
}
\concept{GSS Next}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
