% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scientsn}
\alias{scientsn}
\title{Happy if son scientist}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientsn
}
\details{
Question B. If you had a son, how would you feel if he wanted to be a scientist—would you feel happy, unhappy, or would you not care one way or the other?
}
\section{Values}{

\itemize{
\item \code{1} happy
\item \code{2} not care
\item \code{3} unhappy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3508/vshow}{GSS Data Explorer page for \code{scientsn}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}

Other Science: 
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
