% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{scifrom}
\alias{scifrom}
\title{Main source of information about science and technology}
\source{
General Social Survey https://gss.norc.org
}
\description{
scifrom
}
\details{
Question We are interested in how people get information about science and technology. Where do you get most of your information about science and technology (– newspapers, magazines, the Internet, books or other printed materials, TV, radio, government agencies, family, friends, colleagues,  or some other source)?
}
\section{Values}{

\itemize{
\item \code{1} newspapers
\item \code{2} magazines
\item \code{3} the internet
\item \code{4} books/other printed material
\item \code{5} tv
\item \code{6} radio
\item \code{7} government agencies
\item \code{8} family
\item \code{9} friends/colleagues
\item \code{10} other
\item \code{11} ted talks
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab C/-/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3315/vshow}{GSS Data Explorer page for \code{scifrom}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}

Other Knowledge: 
\code{\link{seeksci}}

Other Science: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
