% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scifromy}
\alias{scifromy}
\title{Where get science and technology info}
\source{
General Social Survey https://gss.norc.org
}
\description{
scifromy
}
\details{
Question We are also interested in how people get information about science and technology. Where do you get most of your information about science and technology printed newspapers, online newspapers, printed magazines, online magazines, other places on the Internet, books or other printed materials, TV, radio, family, friends, colleagues, or some other source?
}
\section{Values}{

\itemize{
\item \code{1} print newspapers
\item \code{2} online newspapers
\item \code{3} print magazines
\item \code{4} online magazines
\item \code{5} other places on the internet
\item \code{6} books/other printed material
\item \code{7} tv
\item \code{8} radio
\item \code{9} family
\item \code{10} friends/colleagues
\item \code{11} other
\item \code{98} don~~t know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3504/vshow}{GSS Data Explorer page for \code{scifromy}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab print newspapers \tab online newspapers \tab print magazines \tab online magazines \tab other places on the internet \tab books/other printed material \tab tv \tab radio \tab family \tab friends/colleagues \tab other \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1386 \tab 47 \tab 50 \tab 51 \tab 16 \tab 123 \tab 58 \tab 239 \tab 15 \tab 19 \tab 19 \tab 11 \tab 7 \tab 3 \tab 2044 \cr
   Total \tab 1386 \tab 47 \tab 50 \tab 51 \tab 16 \tab 123 \tab 58 \tab 239 \tab 15 \tab 19 \tab 19 \tab 11 \tab 7 \tab 3 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
