% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_376.R
\name{sciimp3}
\alias{sciimp3}
\title{How important: conclusion based on solid evidence}
\source{
General Social Survey https://gss.norc.org
}
\description{
sciimp3
}
\details{
Question Now I'm going to read you some statements about science and scientists. Please look at Card B19. How important are each of the following in making something scientific?
C. The conclusions are based on solid evidence. (Would you say this is very important, pretty important, not too important, or not important at all in making something scientific?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} pretty important
\item \verb{[3]} not too important
\item \verb{[4]} not important at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3400/vshow}{GSS Data Explorer page for \code{sciimp3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab pretty important \tab not too important \tab not important at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1474 \tab 277 \tab 36 \tab 20 \tab 51 \tab 6 \tab 4510 \cr
   Total \tab 2646 \tab 1474 \tab 277 \tab 36 \tab 20 \tab 51 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{histsci}},
\code{\link{physcsci}},
\code{\link{sciimp2}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{socsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
