% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_382.R
\name{sciintro}
\alias{sciintro}
\title{Which version of science introduction was used}
\source{
General Social Survey https://gss.norc.org
}
\description{
sciintro
}
\details{
Question Which version of science introduction was used?
}
\section{Values}{

\itemize{
\item \code{1} intro version 1
\item \code{2} intro version 2
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3462/vshow}{GSS Data Explorer page for \code{sciintro}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intenvir}},
\code{\link{intmil}},
\code{\link{intspace}},
\code{\link{scimath}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
