% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_382.R
\name{scimath}
\alias{scimath}
\title{Quality of science & math educ in us is inadequate}
\source{
General Social Survey https://gss.norc.org
}
\description{
scimath
}
\details{
Question The quality of science and mathematics education in American schools is inadequate. (Do you strongly agree, agree, disagree, or strongly disagree?)
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3468/vshow}{GSS Data Explorer page for \code{scimath}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intenvir}},
\code{\link{intmil}},
\code{\link{intspace}},
\code{\link{sciintro}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
