% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{scimode}
\alias{scimode}
\title{Was saq self administered or interview administered}
\source{
General Social Survey https://gss.norc.org
}
\description{
scimode
}
\details{
Question Is the SAQ self administered or interview administered?
}
\section{Values}{

\itemize{
\item \verb{[1]} assigned and done as interviewed
\item \verb{[2]} assigned and done as saq
\item \verb{[3]} assigned as saq, done in interview due to r's request/inability to do as saq
\item \verb{[4]} assigned as saq, done as interview on phone
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3451/vshow}{GSS Data Explorer page for \code{scimode}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab assigned and done as interviewed \tab assigned and done as saq \tab assigned as saq, done in interview due to r's request/inability to do as saq \tab assigned as saq, done as interview on phone \tab Total \cr
   2008 \tab 518 \tab 1237 \tab 140 \tab 89 \tab 39 \tab 2023 \cr
   Total \tab 518 \tab 1237 \tab 140 \tab 89 \tab 39 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
