% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{scimoral}
\alias{scimoral}
\title{Science breaks down ideas of right & wrong}
\source{
General Social Survey https://gss.norc.org
}
\description{
scimoral
}
\details{
Question Here are some things that have been said about science. Would you tell me if you tend to agree or disagree with them?
D. One of the bad effects of science is that it breaks down people''s ideas of right and wrong.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1048/vshow}{GSS Data Explorer page for \code{scimoral}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 474 \tab 961 \tab 29 \tab 17 \tab 1481 \cr
   Total \tab 0 \tab 474 \tab 961 \tab 29 \tab 17 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{scichng}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{switched}}
}
\concept{Religion}
\keyword{variable}
