% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scinews1}
\alias{scinews1}
\title{Newspaper printed or online}
\source{
General Social Survey https://gss.norc.org
}
\description{
scinews1
}
\details{
Question You said you get most of your information about science and technology from newspapers. would that be printed newspapers or online newspapers?
}
\section{Values}{

\itemize{
\item \code{1} printed newspapers
\item \code{2} online newspapers
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016, 2018 \cr
   B/C \tab 2012, 2014 \cr
   C/- \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3500/vshow}{GSS Data Explorer page for \code{scinews1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab printed newspapers \tab other \tab online newspapers \tab no answer \tab Total \cr
   2010 \tab 2028 \tab 15 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 1914 \tab 44 \tab 3 \tab 8 \tab 5 \tab 1974 \cr
   2014 \tab 2455 \tab 68 \tab - \tab 13 \tab 2 \tab 2538 \cr
   2016 \tab 2800 \tab 51 \tab 2 \tab 10 \tab 4 \tab 2867 \cr
   2018 \tab 2298 \tab 31 \tab 3 \tab 15 \tab 1 \tab 2348 \cr
   Total \tab 11495 \tab 209 \tab 9 \tab 46 \tab 12 \tab 11771 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
