% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scinews2}
\alias{scinews2}
\title{Magazine printed or online}
\source{
General Social Survey https://gss.norc.org
}
\description{
scinews2
}
\details{
Question You said you get most of your information about science and technology from magazines. would that be printed magazines or online magazines?
}
\section{Values}{

\itemize{
\item \code{1} printed magazines
\item \code{2} online magazines
\item \code{3} other (specify)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016, 2018 \cr
   B/C \tab 2012, 2014 \cr
   C/- \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3501/vshow}{GSS Data Explorer page for \code{scinews2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab printed magazines \tab online magazines \tab no answer \tab other (specify) \tab Total \cr
   2010 \tab 2018 \tab 23 \tab 2 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 1894 \tab 62 \tab 8 \tab 4 \tab 6 \tab 1974 \cr
   2014 \tab 2456 \tab 65 \tab 14 \tab 2 \tab 1 \tab 2538 \cr
   2016 \tab 2781 \tab 69 \tab 9 \tab 4 \tab 4 \tab 2867 \cr
   2018 \tab 2299 \tab 37 \tab 9 \tab 1 \tab 2 \tab 2348 \cr
   Total \tab 11448 \tab 256 \tab 42 \tab 12 \tab 13 \tab 11771 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
