% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scinews3}
\alias{scinews3}
\title{Where online get info}
\source{
General Social Survey https://gss.norc.org
}
\description{
scinews3
}
\details{
Question You said you get most of your information about science and technology from the Internet. What is the place you are most likely to go on the Internet for science and technology information—online newspapers, online magazines, or some other place on the Internet?
you most likely to go on the Internet for science and technology information?
}
\section{Values}{

\itemize{
\item \code{1} online newspapers
\item \code{2} online magazines
\item \code{3} science site
\item \code{4} news site
\item \code{5} electronic books & reports
\item \code{6} wikipedia
\item \code{7} government site
\item \code{8} social media (facebook, youtube, twitter)
\item \code{10} other
\item \code{11} search engine (google, bing)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab C/- \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3502/vshow}{GSS Data Explorer page for \code{scinews3}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
