% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{scipry}
\alias{scipry}
\title{Scientists pry into things they oughtn't}
\source{
General Social Survey https://gss.norc.org
}
\description{
scipry
}
\details{
Question Here are some things that have been said about science. Would you tell me if you tend to agree or disagree with them?
C. Scientists always seem to be prying into things that they really ought to stay out of.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1047/vshow}{GSS Data Explorer page for \code{scipry}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1988 \tab - \tab 504 \tab 942 \tab 19 \tab 16 \tab - \tab 1481 \cr
   Total \tab 0 \tab 504 \tab 942 \tab 19 \tab 16 \tab 0 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{actchurh}},
\code{\link{afterlif}},
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{ancestrs}},
\code{\link{artgod}},
\code{\link{attend}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{beausprt}},
\code{\link{believe}},
\code{\link{bible}},
\code{\link{bible1}},
\code{\link{bibley}},
\code{\link{blessngs}},
\code{\link{blkwhite}},
\code{\link{bmitzvah}},
\code{\link{bwfriend}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{cathid}},
\code{\link{cathsch}},
\code{\link{cathtemp}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{charisma}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{chrissch}},
\code{\link{churchtx}},
\code{\link{churhact}},
\code{\link{churhgrp}},
\code{\link{churhmem}},
\code{\link{churhpow}},
\code{\link{churhsch}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{closrgod}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{coden_1222}},
\code{\link{colmslm}},
\code{\link{comfort}},
\code{\link{conchurh}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{conorelgsp}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{creation}},
\code{\link{creator}},
\code{\link{crtrheal}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denkid}},
\code{\link{denom}},
\code{\link{denom16}},
\code{\link{denom16_7322}},
\code{\link{denom_7222}},
\code{\link{denomsch}},
\code{\link{densp}},
\code{\link{devil}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{egomeans}},
\code{\link{entity}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{fatalism}},
\code{\link{father}},
\code{\link{feclergy}},
\code{\link{feelgod}},
\code{\link{feelrel}},
\code{\link{fepriest}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{friday}},
\code{\link{friend}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndking}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{fund}},
\code{\link{fund16}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godchnge}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlove}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{godwatch}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{harmony}},
\code{\link{harsexcl}},
\code{\link{healer}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homesch}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{idols}},
\code{\link{impgod}},
\code{\link{implives}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{jew}},
\code{\link{jew16}},
\code{\link{jewtemp}},
\code{\link{join16}},
\code{\link{joylifts}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kid5up}},
\code{\link{king}},
\code{\link{knwclsex}},
\code{\link{lapsed}},
\code{\link{lesspain}},
\code{\link{liberatr}},
\code{\link{libmslm}},
\code{\link{likediff}},
\code{\link{lover}},
\code{\link{maattend}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{madenkid}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{majwoth}},
\code{\link{makefrnd}},
\code{\link{maoth}},
\code{\link{mapa}},
\code{\link{marelig}},
\code{\link{marelkid}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{mediarel}},
\code{\link{meditate}},
\code{\link{memchurh}},
\code{\link{mhgod}},
\code{\link{mindbody}},
\code{\link{miracles}},
\code{\link{monday}},
\code{\link{mother}},
\code{\link{mslmtemp}},
\code{\link{myfaith}},
\code{\link{mywaygod}},
\code{\link{neargod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{norelgsp16}},
\code{\link{notthink}},
\code{\link{numbwfrd}},
\code{\link{numclerg}},
\code{\link{numcong}},
\code{\link{numdays}},
\code{\link{numfrend}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{obeytch}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{oth16}},
\code{\link{othchrch}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{othjew16}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{othsch}},
\code{\link{ownclerg}},
\code{\link{ownfate}},
\code{\link{paattend}},
\code{\link{paden}},
\code{\link{padenkid}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{palefull}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{permoral}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{popespks}},
\code{\link{popespky}},
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf5}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}},
\code{\link{postlife}},
\code{\link{pray}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{prayfreq}},
\code{\link{preach}},
\code{\link{predeter}},
\code{\link{privpray}},
\code{\link{privsch}},
\code{\link{prottemp}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{rdeemlib}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{redeemer}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relcand}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrade}},
\code{\link{relgrpeq}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{relig}},
\code{\link{relig16}},
\code{\link{religcon}},
\code{\link{religid}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{reliten}},
\code{\link{rellife}},
\code{\link{relmarry}},
\code{\link{relmeet}},
\code{\link{relneg}},
\code{\link{relobjct}},
\code{\link{relpersn}},
\code{\link{relsch}},
\code{\link{relsprt}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{reltrad}},
\code{\link{reltrad16}},
\code{\link{reltruth}},
\code{\link{restact}},
\code{\link{rightpeo}},
\code{\link{rosaries}},
\code{\link{rotapple}},
\code{\link{rspctrel}},
\code{\link{saturday}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{schlpray}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scisolve}},
\code{\link{scitesty}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{socright}},
\code{\link{spattend}},
\code{\link{spden}},
\code{\link{spden16}},
\code{\link{spden_7322}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjoth16}},
\code{\link{spkmslm}},
\code{\link{spnorelgsp}},
\code{\link{spoth16}},
\code{\link{spother}},
\code{\link{spothjew}},
\code{\link{spouse}},
\code{\link{sprel}},
\code{\link{sprel16}},
\code{\link{sprtprsn}},
\code{\link{sunday}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{synmem}},
\code{\link{theism}},
\code{\link{thursday}},
\code{\link{tithing}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{tuesday}},
\code{\link{tvrelig}},
\code{\link{typchurh}},
\code{\link{uniongod}},
\code{\link{vistholy}},
\code{\link{wednesdy}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}},
\code{\link{world3}},
\code{\link{worship}},
\code{\link{wrkyoung}}

Other Knowledge: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{anheat}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{boyorgrl}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{electron}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{governor}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{h2olife}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{nanoben}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{poleff20}},
\code{\link{radioact}},
\code{\link{schlhead}},
\code{\link{scifrom}},
\code{\link{scimoral}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{seeksci}},
\code{\link{seesand}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{upbreath}},
\code{\link{usrep}},
\code{\link{viruses}},
\code{\link{weighing}},
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Science: 
\code{\link{accntsci}},
\code{\link{advfront}},
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{artseals}},
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{bettrlfe}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{biosci}},
\code{\link{bizbstgm}},
\code{\link{bizbstgw}},
\code{\link{bizinfgm}},
\code{\link{bizinfgw}},
\code{\link{boyorgrl}},
\code{\link{caremost}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{comorsci}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{eatgm}},
\code{\link{econsci}},
\code{\link{electron}},
\code{\link{engnrsci}},
\code{\link{erosion}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{extinct}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{gmbiz}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{goldfish}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{h2olife}},
\code{\link{harmgood}},
\code{\link{histsci}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{intspace}},
\code{\link{inuit}},
\code{\link{inuitway}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{leadsci}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{medagrgm}},
\code{\link{medagrsc}},
\code{\link{medbstgm}},
\code{\link{medbstsc}},
\code{\link{medinfgm}},
\code{\link{medinfsc}},
\code{\link{medsci}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{newsfrmy}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{noicecap}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{physcsci}},
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}},
\code{\link{polbstgm}},
\code{\link{polbstgw}},
\code{\link{polbstsc}},
\code{\link{polinfgm}},
\code{\link{polinfgw}},
\code{\link{polinfsc}},
\code{\link{radioact}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{salsci}},
\code{\link{salth2o}},
\code{\link{sciagrgw}},
\code{\link{scibnfts}},
\code{\link{scibstgw}},
\code{\link{scichng}},
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientda}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientsn}},
\code{\link{scifaith}},
\code{\link{scifrom}},
\code{\link{scifromy}},
\code{\link{scigrn}},
\code{\link{sciimp1}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{sciinfgo}},
\code{\link{sciinfgw}},
\code{\link{scimath}},
\code{\link{scimoral}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{scisolve}},
\code{\link{scispec}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{sciworse}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}},
\code{\link{scresrch}},
\code{\link{sealevel}},
\code{\link{seeksci}},
\code{\link{seeksciy}},
\code{\link{seesand}},
\code{\link{socsci}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{toofast}},
\code{\link{trustsci}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}},
\code{\link{upbreath}},
\code{\link{viruses}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{weighing}},
\code{\link{whichsci}}
}
\concept{Knowledge}
\concept{Religion}
\concept{Science}
\keyword{variable}
