% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{scisolve}
\alias{scisolve}
\title{Science will solve social problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
scisolve
}
\details{
Question Here are some things that have been said about science. Would you tell me if you tend to agree or disagree with them?
A. Science will solve our social problems like crime and mental illness.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1045/vshow}{GSS Data Explorer page for \code{scisolve}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 334 \tab 1111 \tab 21 \tab 15 \tab 1481 \cr
   Total \tab 0 \tab 334 \tab 1111 \tab 21 \tab 15 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{switched}}
}
\concept{Religion}
\keyword{variable}
