% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{scistudy}
\alias{scistudy}
\title{R has clear understanding of scientific study}
\source{
General Social Survey https://gss.norc.org
}
\description{
scistudy
}
\details{
Question Now, for a slightly different type of question. When you read news stories, you see certain sets of words and terms. We are interested in how many people recognize certain kinds of terms. First, some articles refer to the results of a scientific study. When you read or hear the term scientific study, do you have a clear understanding of what it means, a general sense of what it means, or little understanding of what it means?
}
\section{Values}{

\itemize{
\item \verb{[1]} clear understanding
\item \verb{[2]} general sense
\item \verb{[3]} little understanding
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3328/vshow}{GSS Data Explorer page for \code{scistudy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab clear understanding \tab general sense \tab little understanding \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 550 \tab 932 \tab 349 \tab 31 \tab 2 \tab 4510 \cr
   2008 \tab 518 \tab 424 \tab 749 \tab 309 \tab 17 \tab 6 \tab 2023 \cr
   2010 \tab 1328 \tab 202 \tab 354 \tab 140 \tab 12 \tab 8 \tab 2044 \cr
   2012 \tab 972 \tab 246 \tab 500 \tab 228 \tab 22 \tab 6 \tab 1974 \cr
   2014 \tab 1299 \tab 338 \tab 626 \tab 265 \tab 8 \tab 2 \tab 2538 \cr
   2016 \tab 1477 \tab 418 \tab 676 \tab 271 \tab 18 \tab 7 \tab 2867 \cr
   2018 \tab 1173 \tab 330 \tab 597 \tab 236 \tab 9 \tab 3 \tab 2348 \cr
   Total \tab 9413 \tab 2508 \tab 4434 \tab 1798 \tab 117 \tab 34 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{odds1}},
\code{\link{scibnfts}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
