% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{scitest3}
\alias{scitest3}
\title{Astrology has some scientific truth}
\source{
General Social Survey https://gss.norc.org
}
\description{
scitest3
}
\details{
Question For each statement below, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
C. Astrology - the study of star signs - has some scientific truth.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4743/vshow}{GSS Data Explorer page for \code{scitest3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 138 \tab 572 \tab 320 \tab 317 \tab 128 \tab 82 \tab 1606 \cr
   1994 \tab 1606 \tab 119 \tab 532 \tab 326 \tab 268 \tab 102 \tab 39 \tab 2992 \cr
   Total \tab 1655 \tab 257 \tab 1104 \tab 646 \tab 585 \tab 230 \tab 121 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{toodifme}}
}
\concept{ISSP Environment}
\keyword{variable}
