% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{scitest5}
\alias{scitest5}
\title{All man-made chems cause cancer}
\source{
General Social Survey https://gss.norc.org
}
\description{
scitest5
}
\details{
Question For each statement below, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
E. All man-made chemicals can cause cancer if you eat enough of them.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4745/vshow}{GSS Data Explorer page for \code{scitest5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 124 \tab 598 \tab 444 \tab 197 \tab 111 \tab 83 \tab 1606 \cr
   1994 \tab 1606 \tab 127 \tab 504 \tab 414 \tab 183 \tab 119 \tab 39 \tab 2992 \cr
   2000 \tab 1541 \tab 107 \tab 430 \tab 381 \tab 171 \tab 121 \tab 66 \tab 2817 \cr
   Total \tab 3196 \tab 358 \tab 1532 \tab 1239 \tab 551 \tab 351 \tab 188 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{toodifme}}
}
\concept{ISSP Environment}
\keyword{variable}
