% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{sciworse}
\alias{sciworse}
\title{Humans change nature for the worse}
\source{
General Social Survey https://gss.norc.org
}
\description{
sciworse
}
\details{
Question How much do you agree or disagree with each of these statements?
C. Any change humans cause in nature - no matter how scientific - is likely to make things worse.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4723/vshow}{GSS Data Explorer page for \code{sciworse}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 80 \tab 375 \tab 348 \tab 543 \tab 87 \tab 66 \tab 58 \tab 1606 \cr
   1994 \tab 1606 \tab 67 \tab 308 \tab 298 \tab 501 \tab 111 \tab 56 \tab 45 \tab 2992 \cr
   Total \tab 1655 \tab 147 \tab 683 \tab 646 \tab 1044 \tab 198 \tab 122 \tab 103 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{anrights}},
\code{\link{grnecon}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scifaith}},
\code{\link{scigrn}}
}
\concept{ISSP Environment}
\keyword{variable}
