% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{screlig}
\alias{screlig}
\title{How well religious leaders understand stem cells}
\source{
General Social Survey https://gss.norc.org
}
\description{
screlig
}
\details{
Question Here is another public issue: research that uses human embryonic stem cells. These are cells from human embryos that can develop into many different types of tissue. Some say that cells taken from human embryos are uniquely valuable in medical research. Others say that the same discoveries can be made without destroying living embryos. Please look at Card B6. (On a scale of 1 to 5, where 1 means "Very Well" and 5 means "Not at All"), how well do the following groups understand stem cell research?
B. Religious leaders
}
\section{Values}{

\itemize{
\item \verb{[1]} very well
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3367/vshow}{GSS Data Explorer page for \code{screlig}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very well \tab 2 \tab 3 \tab 4 \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 50 \tab 75 \tab 241 \tab 267 \tab 237 \tab 55 \tab 3 \tab 4510 \cr
   2010 \tab 1569 \tab 26 \tab 29 \tab 113 \tab 153 \tab 122 \tab 28 \tab 4 \tab 2044 \cr
   Total \tab 5151 \tab 76 \tab 104 \tab 354 \tab 420 \tab 359 \tab 83 \tab 7 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgw}},
\code{\link{bizinfgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
