% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{secdocs}
\alias{secdocs}
\title{Does govt classify too many documents as secrets}
\source{
General Social Survey https://gss.norc.org
}
\description{
secdocs
}
\details{
Question Given the world situation, the government protects too many documents by classifying them as SECRET and TOP SECRET. Do you . . .
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5206/vshow}{GSS Data Explorer page for \code{secdocs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 196 \tab 603 \tab 244 \tab 262 \tab 65 \tab 94 \tab 10 \tab 2992 \cr
   1996 \tab 1444 \tab 199 \tab 605 \tab 213 \tab 289 \tab 59 \tab 90 \tab 5 \tab 2904 \cr
   1998 \tab 1445 \tab 225 \tab 515 \tab 248 \tab 216 \tab 53 \tab 113 \tab 17 \tab 2832 \cr
   Total \tab 4407 \tab 620 \tab 1723 \tab 705 \tab 767 \tab 177 \tab 297 \tab 32 \tab 8728 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{sectech}}
}
\concept{National Security}
\keyword{variable}
