% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{secjob}
\alias{secjob}
\title{Importance of job security in a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
secjob
}
\details{
Question On the following list there are various aspects of jobs. Please circle one number to show how important you personally consider it is in a job:
A. Job security?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} neither important nor unimportant
\item \code{4} not important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4507/vshow}{GSS Data Explorer page for \code{secjob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{bosswrks}},
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{hiinc}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{strngun}}
}
\concept{Work Orientation}
\keyword{variable}
