% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{secmilop}
\alias{secmilop}
\title{Government maintain secrecy on military operations}
\source{
General Social Survey https://gss.norc.org
}
\description{
secmilop
}
\details{
Question Should or should not the government maintain a high level of secrecy surrounding: (READ A-D)
B. Military operations
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should
\item \verb{[2]} probably should
\item \verb{[3]} probably should not
\item \verb{[4]} definitely should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5220/vshow}{GSS Data Explorer page for \code{secmilop}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 842 \tab 425 \tab 97 \tab 44 \tab 50 \tab 2 \tab 2904 \cr
   1998 \tab 1445 \tab 736 \tab 479 \tab 72 \tab 30 \tab 59 \tab 11 \tab 2832 \cr
   Total \tab 2889 \tab 1578 \tab 904 \tab 169 \tab 74 \tab 109 \tab 13 \tab 5736 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secprvcy}},
\code{\link{secterr}}
}
\concept{National Security}
\keyword{variable}
