% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{secondwk}
\alias{secondwk}
\title{R has job other than main}
\source{
General Social Survey https://gss.norc.org
}
\description{
secondwk
}
\details{
Question Do you have any jobs besides your main job or do any other work for pay?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2773/vshow}{GSS Data Explorer page for \code{secondwk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 306 \tab 1471 \tab 6 \tab 13 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 285 \tab 1438 \tab 2 \tab 9 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 193 \tab 970 \tab - \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 182 \tab 1061 \tab - \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 261 \tab 1154 \tab 2 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 286 \tab 1679 \tab 4 \tab 14 \tab 15 \tab 3544 \cr
   Total \tab 8336 \tab 1513 \tab 7773 \tab 14 \tab 98 \tab 15 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{hrsrelax}},
\code{\link{learnnew}},
\code{\link{whywkhme}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}}
}
\concept{Quality of Working Life}
\keyword{variable}
