% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{secprvcy}
\alias{secprvcy}
\title{Govt favor secrecy over workers privacy?}
\source{
General Social Survey https://gss.norc.org
}
\description{
secprvcy
}
\details{
Question When faced with the conflict between an individual''s right to privacy and the government''s need to collect personal
background information before giving a SECRET or TOP SECRET clearance, the government should:
\enumerate{
\item Favor protecting an individual''s right to privacy by not gathering information.
\item Favor the government''s need to protect security by gathering information.
}
}
\section{Values}{

\itemize{
\item \verb{[1]} favor indiv right
\item \verb{[2]} favor govt right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5218/vshow}{GSS Data Explorer page for \code{secprvcy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab favor indiv right \tab favor govt right \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 229 \tab 1156 \tab 79 \tab 10 \tab 2992 \cr
   Total \tab 1518 \tab 229 \tab 1156 \tab 79 \tab 10 \tab 2992 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secterr}}
}
\concept{National Security}
\keyword{variable}
