% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{secterr}
\alias{secterr}
\title{Government maintain secrecy on domestic terrorism}
\source{
General Social Survey https://gss.norc.org
}
\description{
secterr
}
\details{
Question Should or should not the government maintain a high level of secrecy surrounding: (READ A-D)
C. Efforts to control domestic terrorism
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5221/vshow}{GSS Data Explorer page for \code{secterr}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 838 \tab 373 \tab 115 \tab 71 \tab 60 \tab 3 \tab 2904 \cr
   1998 \tab 1445 \tab 694 \tab 427 \tab 114 \tab 68 \tab 73 \tab 11 \tab 2832 \cr
   Total \tab 2889 \tab 1532 \tab 800 \tab 229 \tab 139 \tab 133 \tab 14 \tab 5736 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secprvcy}}
}
\concept{National Security}
\keyword{variable}
