% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{sector}
\alias{sector}
\title{Type of college respondent attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
sector
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
A. Sector of Institution?
}
\section{Values}{

\itemize{
\item \code{0} administrative unit
\item \code{1} public, 4-year or above
\item \code{2} private not-for-profit, 4-year or above
\item \code{3} private for-profit, 4-year or above
\item \code{4} public, 2-year
\item \code{5} private not-for-profit, 2-year or less
\item \code{6} private for-profit, 2-year
\item \code{7} public, less than 2-year
\item \code{9} private for-profit, less than 2-year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/69/vshow}{GSS Data Explorer page for \code{sector}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab administrative unit \tab public, 4-year or above \tab private not-for-profit, 4-year or above \tab private for-profit, 4-year or above \tab public, 2-year \tab private not-for-profit, 2-year or less \tab private for-profit, 2-year \tab private for-profit, less than 2-year \tab skipped on web \tab Total \cr
   2012 \tab 1346 \tab 1 \tab 325 \tab 183 \tab 8 \tab 97 \tab 4 \tab 4 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1748 \tab - \tab 432 \tab 242 \tab 23 \tab 90 \tab 2 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 1936 \tab - \tab 538 \tab 269 \tab 15 \tab 100 \tab - \tab 2 \tab - \tab 7 \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 7378 \tab 1 \tab 1295 \tab 694 \tab 46 \tab 287 \tab 6 \tab 7 \tab 6 \tab 7 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
