% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{seefilm}
\alias{seefilm}
\title{Last week went to see a film?}
\source{
General Social Survey https://gss.norc.org
}
\description{
seefilm
}
\details{
Question Now I'm going to ask you about things you did during the last seven days. I'm only interested in what you did during the last seven days.
From last (DAY OF WEEK) to today did you…
C. Go to a movie theater to see a film?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5308/vshow}{GSS Data Explorer page for \code{seefilm}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1953 \tab 131 \tab 819 \tab 1 \tab 2904 \cr
   1998 \tab 1387 \tab 193 \tab 1247 \tab 5 \tab 2832 \cr
   Total \tab 3340 \tab 324 \tab 2066 \tab 6 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attrelig}},
\code{\link{eatout}},
\code{\link{godoc}},
\code{\link{monday}},
\code{\link{numdays}},
\code{\link{sunday}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
