% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{seeksciy}
\alias{seeksciy}
\title{Where get science issue news}
\source{
General Social Survey https://gss.norc.org
}
\description{
seeksciy
}
\details{
Question If you wanted to learn about scientific issues such as global warming or biotechnology, where would you get information – printed newspapers, online newspapers, printed magazines, online magazines, other places on the Internet, books or other printed materials, TV, radio, family, friends, colleagues, or some other source)?
PROBE: What would be your main source of information?
}
\section{Values}{

\itemize{
\item \code{1} print newspapers
\item \code{2} online newspapers
\item \code{3} print magazines
\item \code{4} online magazines
\item \code{5} other places on the internet
\item \code{6} books/other printed material
\item \code{7} tv
\item \code{8} radio
\item \code{9} family
\item \code{10} friends/colleagues
\item \code{11} other
\item \code{98} don~~t know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3505/vshow}{GSS Data Explorer page for \code{seeksciy}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
