% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_150.R
\name{seemovie}
\alias{seemovie}
\title{Went to a movie in lst yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
seemovie
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
L. Go out to see a movie in a theater.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1416/vshow}{GSS Data Explorer page for \code{seemovie}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 1114 \tab 480 \tab 1 \tab 11 \tab 1606 \cr
   1998 \tab 1387 \tab 952 \tab 481 \tab - \tab 12 \tab 2832 \cr
   2002 \tab 1393 \tab 993 \tab 365 \tab 8 \tab 6 \tab 2765 \cr
   Total \tab 2780 \tab 3059 \tab 1326 \tab 9 \tab 29 \tab 7203 \cr
}
}

\seealso{
Other Culture: 
\code{\link{camping}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{huntfish}},
\code{\link{perform}},
\code{\link{plymusic}},
\code{\link{usevcr}}
}
\concept{Culture}
\keyword{variable}
