% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{seenmntl}
\alias{seenmntl}
\title{You or anyone you know seen a psychiatrist}
\source{
General Social Survey https://gss.norc.org
}
\description{
seenmntl
}
\details{
Question Have you or has anyone else you know ever seen a psychiatrist, psychologist, or counselor?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2167/vshow}{GSS Data Explorer page for \code{seenmntl}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 808 \tab 561 \tab 16 \tab 2 \tab 2832 \cr
   Total \tab 1445 \tab 808 \tab 561 \tab 16 \tab 2 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}}
}
\concept{Mental Health}
\keyword{variable}
