% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{seesand}
\alias{seesand}
\title{Which picture shows what she should use for the 2nd pot}
\source{
General Social Survey https://gss.norc.org
}
\description{
seesand
}
\details{
Question Please look at Card I-4. A gardener has an idea that a plant needs sand in the soil for healthy growth. In order to test her idea she uses two pots of plants. She sets up one pot of plants as shown on the top part of the card. Which one of the pictures on the bottom part of the card shows what should she use for the second pot?
}
\section{Values}{

\itemize{
\item \code{1} sunlight. sand and water
\item \code{2} dark cupboard. sand, soil, and water
\item \code{3} dark cupboard. soil and water
\item \code{4} sunlight. sand and soil
\item \code{5} sunlight. soil and water
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3484/vshow}{GSS Data Explorer page for \code{seesand}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{upbreath}},
\code{\link{weighing}}

Other Knowledge: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{upbreath}},
\code{\link{weighing}}

Other Science: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
