% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_223.R
\name{seetalk5}
\alias{seetalk5}
\title{How frequently r sees or talks to person 5}
\source{
General Social Survey https://gss.norc.org
}
\description{
seetalk5
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very often, how frequently do you see or talk to (NAME 5) these days?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - not at all
\item \verb{[10]} 10 - very often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7119/vshow}{GSS Data Explorer page for \code{seetalk5}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all \tab 10 - very often \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2245 \tab 25 \tab 9 \tab 14 \tab 11 \tab 11 \tab 7 \tab 5 \tab 9 \tab 2 \tab 2 \tab 2 \tab 6 \tab 2348 \cr
   Total \tab 2245 \tab 25 \tab 9 \tab 14 \tab 11 \tab 11 \tab 7 \tab 5 \tab 9 \tab 2 \tab 2 \tab 2 \tab 6 \tab 2348 \cr
}
}

\keyword{variable}
