% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{sei10educ}
\alias{sei10educ}
\title{Percentage of some college educ in occ10 based on acs 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
sei10educ
}
\details{
Question Percentage of those who had some college or more education among those who had the same occupation as in OCC10 in American Community Survey between 2010 and 2012 (25-64 years old only). This percentage is not from the GSS but it is used to calculate SEI10.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} i don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{coisco08}},
\code{\link{firstsei}},
\code{\link{masei}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10inc}},
\code{\link{spsei}}
}
\concept{Occupation}
\keyword{variable}
