% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{selfemp1}
\alias{selfemp1}
\title{Employee have more job security than the self-employed}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfemp1
}
\details{
Question To what extent do you agree or disagree with the following statements?
A. Employees have more job security than the self-employed
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4606/vshow}{GSS Data Explorer page for \code{selfemp1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 236 \tab 549 \tab 343 \tab 309 \tab 58 \tab 20 \tab 3 \tab 4510 \cr
   Total \tab 2992 \tab 236 \tab 549 \tab 343 \tab 309 \tab 58 \tab 20 \tab 3 \tab 4510 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{econsup}},
\code{\link{rimpskls}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}
}
\concept{Work Orientation}
\keyword{variable}
