% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{selfems}
\alias{selfems}
\title{(#Selfemsfill)}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfems
}
\details{
Question
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7747/vshow}{GSS Data Explorer page for \code{selfems}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2022 \tab 3150 \tab 116 \tab 278 \tab 3544 \cr
   Total \tab 3150 \tab 116 \tab 278 \tab 3544 \cr
}
}

\keyword{variable}
