% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{selfish}
\alias{selfish}
\title{R is a selfish person}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfish
}
\details{
Question For each of the following, please indicate how well the description applies to you by circling one number:
E. A selfish person. Is this ……
}
\section{Values}{

\itemize{
\item \code{1} a very good description
\item \code{2} a good description
\item \code{3} a fair description
\item \code{4} not a very good description
\item \code{5} not a good description at all
\item \code{9} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3051/vshow}{GSS Data Explorer page for \code{selfish}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a very good description \tab a good description \tab a fair description \tab not a very good description \tab not a good description at all \tab don't know \tab no answer \tab Total \cr
   2004 \tab 432 \tab 23 \tab 64 \tab 207 \tab 1160 \tab 918 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 432 \tab 23 \tab 64 \tab 207 \tab 1160 \tab 918 \tab 1 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{owndoing}},
\code{\link{sadblue}}
}
\concept{Genes and the Environment}
\keyword{variable}
