% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{selfless}
\alias{selfless}
\title{R feels like a selfless caring for others}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfless
}
\details{
Question The following are things that you may experience in your daily life. Please tell me how often these occur.
A. I feel a selfless caring for others.
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never or almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2888/vshow}{GSS Data Explorer page for \code{selfless}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never or almost never \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 128 \tab 177 \tab 266 \tab 328 \tab 288 \tab 136 \tab 42 \tab 7 \tab 2765 \cr
   2004 \tab 1472 \tab 191 \tab 392 \tab 353 \tab 229 \tab 101 \tab 43 \tab 15 \tab 16 \tab 2812 \cr
   2012 \tab 666 \tab 158 \tab 237 \tab 266 \tab 261 \tab 175 \tab 169 \tab 30 \tab 12 \tab 1974 \cr
   2014 \tab 1264 \tab 219 \tab 237 \tab 250 \tab 217 \tab 180 \tab 150 \tab 12 \tab 9 \tab 2538 \cr
   Total \tab 4795 \tab 696 \tab 1043 \tab 1135 \tab 1035 \tab 744 \tab 498 \tab 99 \tab 44 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selffrst}}
}
\concept{Altruism}
\keyword{variable}
