% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{sellbaby}
\alias{sellbaby}
\title{Surrogate mothers should be permitted}
\source{
General Social Survey https://gss.norc.org
}
\description{
sellbaby
}
\details{
Question Recently, some married couples who are unable to have children have paid women, called "surrogate mothers," to bear a child for them. When the child is born, the couple becomes its adoptive parents and the surrogate mother receives a fee. Do you think that this practice should be permitted or forbidden under the law?
}
\section{Values}{

\itemize{
\item \code{1} forbid it
\item \code{2} permit it
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1870/vshow}{GSS Data Explorer page for \code{sellbaby}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab forbid it \tab permit it \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 465 \tab 872 \tab 104 \tab 3 \tab 2904 \cr
   Total \tab 1460 \tab 465 \tab 872 \tab 104 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
