% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{sellorgn}
\alias{sellorgn}
\title{Should people be permitted to sell a kidney}
\source{
General Social Survey https://gss.norc.org
}
\description{
sellorgn
}
\details{
Question A body organ that is much in need and that people may contribute are kidneys. Most people can live with only one kidney, though their chances of survival are better if they have two. Do you believe that people with two healthy kidneys should be permitted to sell a kidney to a hospital or organ center to use for transplants?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely not
\item \verb{[2]} probably not
\item \verb{[3]} perhaps
\item \verb{[4]} probably
\item \verb{[5]} definitely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1868/vshow}{GSS Data Explorer page for \code{sellorgn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab definitely not \tab probably not \tab perhaps \tab probably \tab definitely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 444 \tab 189 \tab 251 \tab 234 \tab 263 \tab 57 \tab 6 \tab 2904 \cr
   Total \tab 1460 \tab 444 \tab 189 \tab 251 \tab 234 \tab 263 \tab 57 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
