% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{semiguns}
\alias{semiguns}
\title{Should semi-auto gun sale be limited to military}
\source{
General Social Survey https://gss.norc.org
}
\description{
semiguns
}
\details{
Question Should semi-automatic, assault weapons or semi-automatic guns known as assault rifles be sold to the general public or should their sales be limited to the military and police?
}
\section{Values}{

\itemize{
\item \code{1} sold to the general public
\item \code{2} sales be limited to the military and police
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3702/vshow}{GSS Data Explorer page for \code{semiguns}}.
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{rifles50}}
}
\concept{Gun Control}
\keyword{variable}
