% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{servegrp}
\alias{servegrp}
\title{Has r served on group committees}
\source{
General Social Survey https://gss.norc.org
}
\description{
servegrp
}
\details{
Question Here is a list of things that members of organizations may participate in. Could you tell me whether or not you have ever done any of these things as a member of (GROUP IN MEMFRAT-MEMNUM OR 332)? (READ EACH ITEM. CIRCLE A NUMBER FOR EACH ACTIVITY).
A. Served on committees
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1015/vshow}{GSS Data Explorer page for \code{servegrp}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{mostact}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
