% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{servpub}
\alias{servpub}
\title{How much the public service serves people}
\source{
General Social Survey https://gss.norc.org
}
\description{
servpub
}
\details{
Question Thinking of the public service in America, how committed is it to serve the people?
}
\section{Values}{

\itemize{
\item \verb{[1]} very committed
\item \verb{[2]} somewhat committed
\item \verb{[3]} not very committed
\item \verb{[4]} not at all committed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8168/vshow}{GSS Data Explorer page for \code{servpub}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab very committed \tab somewhat committed \tab not very committed \tab not at all committed \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 261 \tab 692 \tab 401 \tab 131 \tab 134 \tab 1666 \tab 6 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 261 \tab 692 \tab 401 \tab 131 \tab 134 \tab 1666 \tab 6 \tab 18 \tab 3309 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{leftrght1}},
\code{\link{poltalk}}
}
\concept{ISSP Citizenship}
\keyword{variable}
