% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{sethours}
\alias{sethours}
\title{Who set working hours}
\source{
General Social Survey https://gss.norc.org
}
\description{
sethours
}
\details{
Question Which of the following statements best describes how your working hours are decided? (By working hours we mean here the times you start and finish work, and not the total hours you work per week or month.)
}
\section{Values}{

\itemize{
\item \verb{[1]} employer decides
\item \verb{[2]} i decide w/limts
\item \verb{[3]} free to decide
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4579/vshow}{GSS Data Explorer page for \code{sethours}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab employer decides \tab i decide w/limts \tab free to decide \tab no answer \tab don't know \tab Total \cr
   1998 \tab 1967 \tab 390 \tab 364 \tab 79 \tab 32 \tab - \tab 2832 \cr
   2006 \tab 3492 \tab 450 \tab 404 \tab 160 \tab 2 \tab 2 \tab 4510 \cr
   Total \tab 5459 \tab 840 \tab 768 \tab 239 \tab 34 \tab 2 \tab 7342 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{wantjob1}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
