% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_304.R
\name{sethrs}
\alias{sethrs}
\title{How are your working hours decided}
\source{
General Social Survey https://gss.norc.org
}
\description{
sethrs
}
\details{
Question Which of the following statements best describes how your working hours are decided? In this question, working hours refers to the total number of hours you work each week, not the time you start and finish work each day.
(HAND CARD L1)
}
\section{Values}{

\itemize{
\item \code{1} the total number of hours i work each week is decided by my employer with little or no input from me
\item \code{2} the total number of hours i work each week is decided by my employer but with my input
\item \code{3} i can decide how many hours i work each week, within certain limits
\item \code{4} i am entirely free to decide how many hours i work each week
\item \code{5} how many hours i work a week depends on things outside of my control and outside of my employer's control
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6295/vshow}{GSS Data Explorer page for \code{sethrs}}.
}

\seealso{
Other Flexible Work: 
\code{\link{advsched}},
\code{\link{leasthrs}}
}
\concept{Flexible Work}
\keyword{variable}
