% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{setprice}
\alias{setprice}
\title{Govt control of prices}
\source{
General Social Survey https://gss.norc.org
}
\description{
setprice
}
\details{
Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it.
B. Control of prices by legislation.
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor
\item \code{2} in favor
\item \code{3} neither
\item \code{4} against
\item \code{5} strongly against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1996 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4050/vshow}{GSS Data Explorer page for \code{setprice}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly in favor \tab in favor \tab neither \tab against \tab strongly against \tab no answer \tab don't know \tab Total \cr
   1985 \tab 857 \tab 57 \tab 194 \tab 132 \tab 204 \tab 78 \tab 12 \tab - \tab 1534 \cr
   1990 \tab 155 \tab 109 \tab 315 \tab 307 \tab 303 \tab 138 \tab 44 \tab 1 \tab 1372 \cr
   1996 \tab 1572 \tab 119 \tab 327 \tab 289 \tab 376 \tab 164 \tab 52 \tab 5 \tab 2904 \cr
   Total \tab 2584 \tab 285 \tab 836 \tab 728 \tab 883 \tab 380 \tab 108 \tab 6 \tab 5810 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{poleff10}},
\code{\link{savejobs}},
\code{\link{setwage}},
\code{\link{spenviro}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
