% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_307.R
\name{setthngs}
\alias{setthngs}
\title{How often r set way things done}
\source{
General Social Survey https://gss.norc.org
}
\description{
setthngs
}
\details{
Question B. How often do you participate with others in helping set the way things are done on your job?
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2794/vshow}{GSS Data Explorer page for \code{setthngs}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 819 \tab 616 \tab 215 \tab 126 \tab 6 \tab 14 \tab 2765 \cr
   2006 \tab 2776 \tab 739 \tab 565 \tab 248 \tab 155 \tab 15 \tab 12 \tab 4510 \cr
   Total \tab 3745 \tab 1558 \tab 1181 \tab 463 \tab 281 \tab 21 \tab 26 \tab 7275 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{partteam}},
\code{\link{prodctiv}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{safehlth}},
\code{\link{toofewwk}},
\code{\link{trdunion}},
\code{\link{wkdecide}},
\code{\link{wksmooth}}
}
\concept{Quality of Working Life}
\keyword{variable}
