% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{sex}
\alias{sex}
\title{Respondents sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
sex
}
\details{
Question CODE RESPONDENT'S SEX
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/81/vshow}{GSS Data Explorer page for \code{sex}}.

Counts by year:\tabular{llllllll}{
   year \tab male \tab female \tab iap \tab no answer \tab skipped on web \tab don't know \tab Total \cr
   1972 \tab 807 \tab 806 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 701 \tab 803 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 691 \tab 793 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 670 \tab 820 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 669 \tab 830 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 693 \tab 837 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 643 \tab 889 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 641 \tab 827 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 779 \tab 1081 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 690 \tab 909 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 598 \tab 875 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 688 \tab 846 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 621 \tab 849 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 778 \tab 1041 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 638 \tab 843 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 660 \tab 877 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 604 \tab 768 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 636 \tab 881 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 685 \tab 921 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1290 \tab 1702 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1285 \tab 1619 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1232 \tab 1600 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1229 \tab 1588 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1228 \tab 1537 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1280 \tab 1532 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2003 \tab 2507 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 930 \tab 1093 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 891 \tab 1153 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 885 \tab 1089 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1141 \tab 1397 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1276 \tab 1591 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1052 \tab 1296 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1736 \tab 2204 \tab 19 \tab 71 \tab 2 \tab - \tab 4032 \cr
   2022 \tab 1627 \tab 1897 \tab - \tab 9 \tab 3 \tab 8 \tab 3544 \cr
   Total \tab 31977 \tab 40301 \tab 19 \tab 80 \tab 5 \tab 8 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
