% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_95.R
\name{sex2}
\alias{sex2}
\title{Sex of person number 2}
\source{
General Social Survey https://gss.norc.org
}
\description{
sex2
}
\details{
Question (NAME) is (male/female)?  Is that correct?  ASK FOR EACH NAME.
2. Sex of Second Person
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/866/vshow}{GSS Data Explorer page for \code{sex2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1985 \tab 364 \tab 606 \tab 559 \tab 5 \tab 1534 \cr
   2004 \tab 1977 \tab 353 \tab 435 \tab 47 \tab 2812 \cr
   Total \tab 2341 \tab 959 \tab 994 \tab 52 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{sex1}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}}
}
\concept{Social Networks}
\keyword{variable}
