% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_95.R
\name{sex4}
\alias{sex4}
\title{Sex of person number 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
sex4
}
\details{
Question (NAME) is (male/female)?  Is that correct?  ASK FOR EACH NAME.
4. Sex of Fourth Person
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/868/vshow}{GSS Data Explorer page for \code{sex4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab not available in this year \tab Total \cr
   1985 \tab 920 \tab 292 \tab 319 \tab 3 \tab - \tab 1534 \cr
   2004 \tab 2472 \tab 137 \tab 157 \tab 46 \tab - \tab 2812 \cr
   Total \tab 3392 \tab 429 \tab 476 \tab 49 \tab 0 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{acqasian}},
\code{\link{acqattnd}},
\code{\link{acqblack}},
\code{\link{acqbrnda}},
\code{\link{acqchild}},
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqdems}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhme}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmno}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqfmwht}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjans}},
\code{\link{acqjose}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqlinda}},
\code{\link{acqmaria}},
\code{\link{acqmark}},
\code{\link{acqmils}},
\code{\link{acqmyrac}},
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhlib}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}},
\code{\link{acqnhwht}},
\code{\link{acqnoatt}},
\code{\link{acqprisn}},
\code{\link{acqrachl}},
\code{\link{acqreps}},
\code{\link{acqshawn}},
\code{\link{acqsocs}},
\code{\link{acqunemp}},
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvalib}},
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhme}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkno}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}},
\code{\link{acqwkwht}},
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{comamind}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{fricall}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndknow}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{grpother}},
\code{\link{grpsprts}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{hlpinflu}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{lentto}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{peocntct}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkedto}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}},
\code{\link{trtasian}},
\code{\link{trtattnd}},
\code{\link{trtblack}},
\code{\link{trtbrnda}},
\code{\link{trtchild}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtcops}},
\code{\link{trtdems}},
\code{\link{trtelecs}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtjose}},
\code{\link{trtkaren}},
\code{\link{trtkeith}},
\code{\link{trtkevin}},
\code{\link{trtlaws}},
\code{\link{trtlib}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtmils}},
\code{\link{trtmyrac}},
\code{\link{trtnoatt}},
\code{\link{trtprisn}},
\code{\link{trtrachl}},
\code{\link{trtreps}},
\code{\link{trtshawn}},
\code{\link{trtsocs}},
\code{\link{trtunemp}},
\code{\link{trtwhite}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{youinflu}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}

Other Sex: 
\code{\link{acpttg}},
\code{\link{acqntsex}},
\code{\link{attractd}},
\code{\link{chngdocs}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{condom}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{family16sex}},
\code{\link{frndsex}},
\code{\link{harsexcl}},
\code{\link{harsexjb}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kidxfilm}},
\code{\link{knwclsex}},
\code{\link{matesex}},
\code{\link{numclerg}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{othersex}},
\code{\link{ownclerg}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{premars1}},
\code{\link{premarsx}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{relgendr}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sellsex}},
\code{\link{sex}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex5}},
\code{\link{sexeduc}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexnow2}},
\code{\link{sexornt}},
\code{\link{sexsex}},
\code{\link{sexsex18}},
\code{\link{sexsex5}},
\code{\link{sexunion}},
\code{\link{spvtrfair}},
\code{\link{teensex}},
\code{\link{teensex1}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{xmarsex}},
\code{\link{xmarsex1}},
\code{\link{xmovie}},
\code{\link{xmovie1}}

Other Social Networks: 
\code{\link{acqasian}},
\code{\link{acqattnd}},
\code{\link{acqblack}},
\code{\link{acqbrnda}},
\code{\link{acqchild}},
\code{\link{acqcohab}},
\code{\link{acqcon}},
\code{\link{acqcops}},
\code{\link{acqdems}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhme}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmno}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}},
\code{\link{acqfmwht}},
\code{\link{acqgay}},
\code{\link{acqgoatt}},
\code{\link{acqhisp}},
\code{\link{acqhome}},
\code{\link{acqjans}},
\code{\link{acqjose}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqlaws}},
\code{\link{acqlib}},
\code{\link{acqlinda}},
\code{\link{acqmaria}},
\code{\link{acqmark}},
\code{\link{acqmils}},
\code{\link{acqmyrac}},
\code{\link{acqnhasn}},
\code{\link{acqnhblk}},
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhgay}},
\code{\link{acqnhgo}},
\code{\link{acqnhhme}},
\code{\link{acqnhhsp}},
\code{\link{acqnhlib}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}},
\code{\link{acqnhno}},
\code{\link{acqnhpri}},
\code{\link{acqnhune}},
\code{\link{acqnhwht}},
\code{\link{acqnoatt}},
\code{\link{acqprisn}},
\code{\link{acqrachl}},
\code{\link{acqreps}},
\code{\link{acqshawn}},
\code{\link{acqsocs}},
\code{\link{acqunemp}},
\code{\link{acqvaasn}},
\code{\link{acqvablk}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvagay}},
\code{\link{acqvago}},
\code{\link{acqvahme}},
\code{\link{acqvahsp}},
\code{\link{acqvalib}},
\code{\link{acqvalin}},
\code{\link{acqvamrk}},
\code{\link{acqvano}},
\code{\link{acqvapri}},
\code{\link{acqvaune}},
\code{\link{acqvawht}},
\code{\link{acqwhite}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkcoh}},
\code{\link{acqwkcon}},
\code{\link{acqwkgay}},
\code{\link{acqwkgo}},
\code{\link{acqwkhme}},
\code{\link{acqwkhsp}},
\code{\link{acqwklib}},
\code{\link{acqwklin}},
\code{\link{acqwkmrk}},
\code{\link{acqwkno}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}},
\code{\link{acqwkwht}},
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{comamind}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{fricall}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{frndknow}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{grpother}},
\code{\link{grpsprts}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{hlpinflu}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{lentto}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam4}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{peocntct}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkedto}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}},
\code{\link{trtasian}},
\code{\link{trtattnd}},
\code{\link{trtblack}},
\code{\link{trtbrnda}},
\code{\link{trtchild}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtcops}},
\code{\link{trtdems}},
\code{\link{trtelecs}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtjose}},
\code{\link{trtkaren}},
\code{\link{trtkeith}},
\code{\link{trtkevin}},
\code{\link{trtlaws}},
\code{\link{trtlib}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtmils}},
\code{\link{trtmyrac}},
\code{\link{trtnoatt}},
\code{\link{trtprisn}},
\code{\link{trtrachl}},
\code{\link{trtreps}},
\code{\link{trtshawn}},
\code{\link{trtsocs}},
\code{\link{trtunemp}},
\code{\link{trtwhite}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{youinflu}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Sex}
\concept{Social Networks}
\keyword{variable}
