% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_95.R
\name{sex5}
\alias{sex5}
\title{Sex of person number 5}
\source{
General Social Survey https://gss.norc.org
}
\description{
sex5
}
\details{
Question (NAME) is (male/female)?  Is that correct?  ASK FOR EACH NAME.
5. Sex of Fifth Person
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/869/vshow}{GSS Data Explorer page for \code{sex5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1985 \tab 1153 \tab 194 \tab 183 \tab 4 \tab 1534 \cr
   2004 \tab 2600 \tab 71 \tab 95 \tab 46 \tab 2812 \cr
   Total \tab 3753 \tab 265 \tab 278 \tab 50 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}}
}
\concept{Social Networks}
\keyword{variable}
