% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{sexhar}
\alias{sexhar}
\title{R experienced sexual harassment}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexhar
}
\details{
Question Sometimes at work people find themselves the object of sexual advances, propositions, or unwanted sexual discussions from co-workers or supervisors. The advances sometimes involve physical contact and sometimes just involve sexual conversations. Has this ever happened to you?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} never have worked
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4416/vshow}{GSS Data Explorer page for \code{sexhar}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{teensex1}}
}
\concept{Gender}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
