% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{sexnow}
\alias{sexnow}
\title{What is your current gender?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexnow
}
\details{
Question What is your current gender?
}
\section{Values}{

\itemize{
\item \verb{[1]} female
\item \verb{[2]} male
\item \verb{[3]} transgender
\item \verb{[4]} a gender not listed here
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7293/vshow}{GSS Data Explorer page for \code{sexnow}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab female \tab male \tab transgender \tab a gender not listed here \tab no answer \tab Total \cr
   2018 \tab 939 \tab 758 \tab 640 \tab 2 \tab 1 \tab 8 \tab 2348 \cr
   Total \tab 939 \tab 758 \tab 640 \tab 2 \tab 1 \tab 8 \tab 2348 \cr
}
}

\keyword{variable}
