% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{sexnow2}
\alias{sexnow2}
\title{R's sex now (3 category)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexnow2
}
\details{
Question Do you describe yourself as male, female, or (REDACTED)?
Question edits are in compliance with federal administration directives released in January 2025.
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[3]} redacted/none of these
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8170/vshow}{GSS Data Explorer page for \code{sexnow2}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab male \tab female \tab redacted/none of these \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1450 \tab 1812 \tab 24 \tab 11 \tab 12 \tab 3309 \cr
   Total \tab 0 \tab 1450 \tab 1812 \tab 24 \tab 11 \tab 12 \tab 3309 \cr
}
}

\seealso{
Other Core: 
\code{\link{conorelgsp}},
\code{\link{family16sex}},
\code{\link{femsee1}},
\code{\link{mascsee1}},
\code{\link{mighist}},
\code{\link{racdif5x}}

Other Sex: 
\code{\link{family16sex}}
}
\concept{Core}
\concept{Sex}
\keyword{variable}
