% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{sexornt}
\alias{sexornt}
\title{Sexual orientation}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexornt
}
\details{
Question Which of the following best describes you?
}
\section{Values}{

\itemize{
\item \verb{[1]} gay, lesbian, or homosexual
\item \verb{[2]} bisexual
\item \verb{[3]} heterosexual or straight
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2010, 2012, 2014, 2024 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5081/vshow}{GSS Data Explorer page for \code{sexornt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab gay, lesbian, or homosexual \tab bisexual \tab heterosexual or straight \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 236 \tab 33 \tab 27 \tab 1699 \tab 7 \tab 21 \tab - \tab 2023 \cr
   2010 \tab 202 \tab 27 \tab 31 \tab 1749 \tab 9 \tab 26 \tab - \tab 2044 \cr
   2012 \tab 237 \tab 27 \tab 43 \tab 1637 \tab 3 \tab 27 \tab - \tab 1974 \cr
   2014 \tab 188 \tab 45 \tab 65 \tab 2195 \tab 6 \tab 39 \tab - \tab 2538 \cr
   2016 \tab 1095 \tab 46 \tab 56 \tab 1641 \tab 5 \tab 24 \tab - \tab 2867 \cr
   2018 \tab 942 \tab 31 \tab 51 \tab 1291 \tab 6 \tab 27 \tab - \tab 2348 \cr
   2021 \tab 1682 \tab 76 \tab 96 \tab 2088 \tab 53 \tab 2 \tab 35 \tab 4032 \cr
   2022 \tab 1900 \tab 53 \tab 82 \tab 1409 \tab 62 \tab 7 \tab 31 \tab 3544 \cr
   2024 \tab - \tab 67 \tab 122 \tab 3052 \tab 41 \tab 11 \tab 16 \tab 3309 \cr
   Total \tab 6482 \tab 405 \tab 573 \tab 16761 \tab 192 \tab 184 \tab 82 \tab 24679 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{frndsex}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{idu30}},
\code{\link{matesex}},
\code{\link{nocondom}},
\code{\link{nodocidu}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexornt1}},
\code{\link{sexpaid}},
\code{\link{sexsex}},
\code{\link{sexsex5}}

Other Gays And Lesbians: 
\code{\link{acqfmgay}},
\code{\link{acqgay}},
\code{\link{acqnhgay}},
\code{\link{acqvagay}},
\code{\link{acqwkgay}},
\code{\link{asksexor}},
\code{\link{attractd}},
\code{\link{colhomo}},
\code{\link{draftgay}},
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{libhomo}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{negjob5}},
\code{\link{sexsex18}},
\code{\link{sexunion}},
\code{\link{spkhomo}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}},
\code{\link{trtgay}}

Other Sex: 
\code{\link{acpttg}},
\code{\link{acqntsex}},
\code{\link{attractd}},
\code{\link{chngdocs}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{condom}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{family16sex}},
\code{\link{frndsex}},
\code{\link{harsexcl}},
\code{\link{harsexjb}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kidxfilm}},
\code{\link{knwclsex}},
\code{\link{matesex}},
\code{\link{numclerg}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{othersex}},
\code{\link{ownclerg}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{premars1}},
\code{\link{premarsx}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{relgendr}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sellsex}},
\code{\link{sex}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sexeduc}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexnow2}},
\code{\link{sexsex}},
\code{\link{sexsex18}},
\code{\link{sexsex5}},
\code{\link{sexunion}},
\code{\link{spvtrfair}},
\code{\link{teensex}},
\code{\link{teensex1}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{xmarsex}},
\code{\link{xmarsex1}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Gays And Lesbians}
\concept{High Risk Behaviors}
\concept{Sex}
\keyword{variable}
