% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{sexornt1}
\alias{sexornt1}
\title{Sexual orientation (reask, self-administered)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexornt1
}
\details{
Question Which of the following best describes you?
}
\section{Values}{

\itemize{
\item \verb{[1]} gay, lesbian, or homosexual
\item \verb{[2]} bisexual
\item \verb{[3]} heterosexual or straight
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8171/vshow}{GSS Data Explorer page for \code{sexornt1}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab gay, lesbian, or homosexual \tab bisexual \tab heterosexual or straight \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 41 \tab 86 \tab 1893 \tab 55 \tab 1186 \tab 5 \tab 43 \tab 3309 \cr
   Total \tab 0 \tab 41 \tab 86 \tab 1893 \tab 55 \tab 1186 \tab 5 \tab 43 \tab 3309 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{frndsex}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{idu30}},
\code{\link{matesex}},
\code{\link{nocondom}},
\code{\link{nodocidu}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexornt}},
\code{\link{sexpaid}},
\code{\link{sexsex}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
