% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{sexpaid}
\alias{sexpaid}
\title{Ever had sex with a person you paid or who paid you for sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexpaid
}
\details{
Question Have you ever had sex with a person  you paid or who paid you for sex?
As of 2024, SEXPAID has been slightly adjusted to help protect confidentiality. Please see the 2024 codebook for more information.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8204/vshow}{GSS Data Explorer page for \code{sexpaid}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 124 \tab 1891 \tab 53 \tab 1187 \tab 10 \tab 44 \tab 3309 \cr
   Total \tab 0 \tab 124 \tab 1891 \tab 53 \tab 1187 \tab 10 \tab 44 \tab 3309 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{nocondom}},
\code{\link{nodocidu}},
\code{\link{sexornt1}}

Other Health: 
\code{\link{nodocidu}}
}
\concept{Health}
\concept{High Risk Behaviors}
\keyword{variable}
