% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{sexsex}
\alias{sexsex}
\title{Sex of sex partner's in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexsex
}
\details{
Question Have your sex partners in the last 12 months been…
}
\section{Values}{

\itemize{
\item \verb{[1]} exclusively male
\item \verb{[2]} both male and female
\item \verb{[3]} exclusively female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5056/vshow}{GSS Data Explorer page for \code{sexsex}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab exclusively male \tab both male and female \tab exclusively female \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1988 \tab 409 \tab 556 \tab 2 \tab 461 \tab 53 \tab - \tab - \tab 1481 \cr
   1989 \tab 444 \tab 558 \tab 3 \tab 498 \tab 34 \tab - \tab - \tab 1537 \cr
   1990 \tab 429 \tab 468 \tab 5 \tab 424 \tab 46 \tab - \tab - \tab 1372 \cr
   1991 \tab 511 \tab 530 \tab 3 \tab 412 \tab 61 \tab - \tab - \tab 1517 \cr
   1993 \tab 427 \tab 592 \tab 6 \tab 504 \tab 77 \tab - \tab - \tab 1606 \cr
   1994 \tab 809 \tab 1108 \tab 8 \tab 953 \tab 114 \tab - \tab - \tab 2992 \cr
   1996 \tab 756 \tab 1063 \tab 18 \tab 950 \tab 117 \tab - \tab - \tab 2904 \cr
   1998 \tab 895 \tab 990 \tab 17 \tab 841 \tab 89 \tab - \tab - \tab 2832 \cr
   2000 \tab 938 \tab 934 \tab 14 \tab 822 \tab 109 \tab - \tab - \tab 2817 \cr
   2002 \tab 1000 \tab 887 \tab 12 \tab 821 \tab 45 \tab - \tab - \tab 2765 \cr
   2004 \tab 1013 \tab 943 \tab 8 \tab 813 \tab 33 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 2692 \tab 967 \tab 20 \tab 809 \tab 22 \tab - \tab - \tab 4510 \cr
   2008 \tab 651 \tab 687 \tab 13 \tab 653 \tab 14 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 632 \tab 716 \tab 10 \tab 667 \tab 18 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 643 \tab 700 \tab 18 \tab 592 \tab 21 \tab - \tab - \tab 1974 \cr
   2014 \tab 727 \tab 931 \tab 20 \tab 820 \tab 38 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 1506 \tab 694 \tab 26 \tab 626 \tab 15 \tab - \tab - \tab 2867 \cr
   2018 \tab 1301 \tab 508 \tab 21 \tab 505 \tab 13 \tab - \tab - \tab 2348 \cr
   2021 \tab 2307 \tab 884 \tab 16 \tab 789 \tab 19 \tab 17 \tab - \tab 4032 \cr
   2022 \tab 2409 \tab 574 \tab 17 \tab 520 \tab 18 \tab 6 \tab - \tab 3544 \cr
   Total \tab 20499 \tab 15290 \tab 257 \tab 13480 \tab 956 \tab 33 \tab 0 \tab 50515 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{frndsex}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{idu30}},
\code{\link{matesex}},
\code{\link{nocondom}},
\code{\link{nodocidu}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexornt}},
\code{\link{sexornt1}},
\code{\link{sexpaid}},
\code{\link{sexsex5}}

Other Sex: 
\code{\link{acpttg}},
\code{\link{acqntsex}},
\code{\link{attractd}},
\code{\link{chngdocs}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{condom}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{family16sex}},
\code{\link{frndsex}},
\code{\link{harsexcl}},
\code{\link{harsexjb}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kidxfilm}},
\code{\link{knwclsex}},
\code{\link{matesex}},
\code{\link{numclerg}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{othersex}},
\code{\link{ownclerg}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{premars1}},
\code{\link{premarsx}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{relgendr}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sellsex}},
\code{\link{sex}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sexeduc}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexnow2}},
\code{\link{sexornt}},
\code{\link{sexsex18}},
\code{\link{sexsex5}},
\code{\link{sexunion}},
\code{\link{spvtrfair}},
\code{\link{teensex}},
\code{\link{teensex1}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{xmarsex}},
\code{\link{xmarsex1}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{High Risk Behaviors}
\concept{Sex}
\keyword{variable}
