% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{sexsex18}
\alias{sexsex18}
\title{Since age 18, have your sex partner's been...}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexsex18
}
\details{
Question Since the age of 18, have your sex partners been:
}
\section{Values}{

\itemize{
\item \verb{[1]} only men
\item \verb{[2]} some women, but mostly men
\item \verb{[3]} equally men and women
\item \verb{[4]} some men, but mostly women
\item \verb{[5]} only women
\item \verb{[6]} i have not had any sex partners
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3906/vshow}{GSS Data Explorer page for \code{sexsex18}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab only men \tab some women, but mostly men \tab equally men and women \tab some men, but mostly women \tab only women \tab i have not had any sex partners \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 239 \tab 890 \tab 36 \tab 7 \tab 13 \tab 754 \tab 31 \tab 3 \tab 50 \tab - \tab 2023 \cr
   2012 \tab 237 \tab 859 \tab 56 \tab 11 \tab 22 \tab 691 \tab 57 \tab - \tab 41 \tab - \tab 1974 \cr
   Total \tab 476 \tab 1749 \tab 92 \tab 18 \tab 35 \tab 1445 \tab 88 \tab 3 \tab 91 \tab 0 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{dwell5}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evkid}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{mykids}},
\code{\link{negjob5}},
\code{\link{othkids}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{sppart}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}}

Other Gays And Lesbians: 
\code{\link{acqfmgay}},
\code{\link{acqgay}},
\code{\link{acqnhgay}},
\code{\link{acqvagay}},
\code{\link{acqwkgay}},
\code{\link{asksexor}},
\code{\link{attractd}},
\code{\link{colhomo}},
\code{\link{draftgay}},
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{libhomo}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{negjob5}},
\code{\link{sexornt}},
\code{\link{sexunion}},
\code{\link{spkhomo}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}},
\code{\link{trtgay}}

Other Sex: 
\code{\link{acpttg}},
\code{\link{acqntsex}},
\code{\link{attractd}},
\code{\link{chngdocs}},
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{clsclsex}},
\code{\link{condom}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{family16sex}},
\code{\link{frndsex}},
\code{\link{harsexcl}},
\code{\link{harsexjb}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{homosex1}},
\code{\link{hushrel1}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kidxfilm}},
\code{\link{knwclsex}},
\code{\link{matesex}},
\code{\link{numclerg}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}},
\code{\link{openrel1}},
\code{\link{openrel2}},
\code{\link{openrel3}},
\code{\link{othersex}},
\code{\link{ownclerg}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{pikupsex}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninf}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{premars1}},
\code{\link{premarsx}},
\code{\link{ptnrsxnow1}},
\code{\link{relatsex}},
\code{\link{relgendr}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sellsex}},
\code{\link{sex}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sexeduc}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexnow2}},
\code{\link{sexornt}},
\code{\link{sexsex}},
\code{\link{sexsex5}},
\code{\link{sexunion}},
\code{\link{spvtrfair}},
\code{\link{teensex}},
\code{\link{teensex1}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{xmarsex}},
\code{\link{xmarsex1}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Gays And Lesbians}
\concept{Sex}
\concept{Sexual Orientation}
\keyword{variable}
