% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_182.R
\name{shakeblu}
\alias{shakeblu}
\title{Felt that couldn't shake the blues?}
\source{
General Social Survey https://gss.norc.org
}
\description{
shakeblu
}
\details{
Question Now I'm going to read a list of different feelings that people sometimes have. After each one, I would like you to tell me on how many days you have felt this way during the past 7 days. On how many days in the past 7 days have you... (SPECIFY NUMBER OF DAYS)
(IF RESPONDENT ASKS, SAY THAT FEELING THIS WAY AT ANY TIME DURING A DAY MAKES THAT DAY "COUNT.")
A. Felt that you couldn't shake the blues?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1706/vshow}{GSS Data Explorer page for \code{shakeblu}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 791 \tab 230 \tab 160 \tab 105 \tab 46 \tab 37 \tab 9 \tab 66 \tab 11 \tab 5 \tab 2904 \cr
   Total \tab 1444 \tab 791 \tab 230 \tab 160 \tab 105 \tab 46 \tab 37 \tab 9 \tab 66 \tab 11 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{ashamed}},
\code{\link{calm}},
\code{\link{hapfeel}},
\code{\link{outraged}},
\code{\link{sad}}
}
\concept{Emotions}
\keyword{variable}
