% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{shop1}
\alias{shop1}
\title{Who in household shops for groceries}
\source{
General Social Survey https://gss.norc.org
}
\description{
shop1
}
\details{
Question In your household who does the following things...?
D. Shops for groceries?
}
\section{Values}{

\itemize{
\item \verb{[1]} always me
\item \verb{[2]} usually me
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually my spouse/partner
\item \verb{[5]} always my spouse/partner
\item \verb{[6]} is done by a third person
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4432/vshow}{GSS Data Explorer page for \code{shop1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab always me \tab usually me \tab about equal or both together \tab usually my spouse/partner \tab always my spouse/partner \tab is done by a third person \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 2058 \tab 138 \tab 119 \tab 290 \tab 131 \tab 5 \tab 1 \tab 1 \tab 22 \tab - \tab 2765 \cr
   2012 \tab 1274 \tab 134 \tab 105 \tab 288 \tab 92 \tab 66 \tab 12 \tab - \tab 3 \tab - \tab 1974 \cr
   2022 \tab 3024 \tab 89 \tab 95 \tab 193 \tab 87 \tab 45 \tab 3 \tab 4 \tab 1 \tab 3 \tab 3544 \cr
   Total \tab 6356 \tab 361 \tab 319 \tab 771 \tab 310 \tab 116 \tab 16 \tab 5 \tab 26 \tab 3 \tab 8283 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{deckids}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{spborn}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
