% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{shopfood}
\alias{shopfood}
\title{Who in household shops for groceries}
\source{
General Social Survey https://gss.norc.org
}
\description{
shopfood
}
\details{
Question In your household who does the following things . . .
D. Shops for groceries.
}
\section{Values}{

\itemize{
\item \verb{[1]} always the woman
\item \verb{[2]} usually the woman
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually the man
\item \verb{[5]} always the man
\item \verb{[6]} is done by a third person
\item \verb{[7]} not married
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4421/vshow}{GSS Data Explorer page for \code{shopfood}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab always the woman \tab usually the woman \tab about equal or both together \tab usually the man \tab always the man \tab is done by a third person \tab not married \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 138 \tab 262 \tab 358 \tab 40 \tab 13 \tab 1 \tab 604 \tab 1 \tab 30 \tab 2992 \cr
   Total \tab 1545 \tab 138 \tab 262 \tab 358 \tab 40 \tab 13 \tab 1 \tab 604 \tab 1 \tab 30 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{repairs1}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
