% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{shoprel}
\alias{shoprel}
\title{Ever shopped around for a church}
\source{
General Social Survey https://gss.norc.org
}
\description{
shoprel
}
\details{
Question As an adult, have you ever shopped around for a church or synagogue--that is, gone to different churches or synagogues or other places of worship in order to compare and decide which one you wanted to attend regularly?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1155/vshow}{GSS Data Explorer page for \code{shoprel}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 561 \tab 870 \tab 1 \tab 13 \tab 2832 \cr
   Total \tab 1387 \tab 561 \tab 870 \tab 1 \tab 13 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{chrissch}},
\code{\link{denomsch}},
\code{\link{implives}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shopnum}}
}
\concept{Religion}
\keyword{variable}
