% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{shortcom}
\alias{shortcom}
\title{World better if america acknowledged shortcomings}
\source{
General Social Survey https://gss.norc.org
}
\description{
shortcom
}
\details{
Question How much do you agree of disagree with the following statements?
A. The world would be a better place if Americans acknowledged America''s shortcomings.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4898/vshow}{GSS Data Explorer page for \code{shortcom}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab 5 \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1264 \tab 222 \tab 676 \tab 235 \tab 83 \tab 12 \tab 43 \tab 3 \tab 2538 \cr
   Total \tab 1264 \tab 222 \tab 676 \tab 235 \tab 83 \tab 12 \tab 43 \tab 3 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}}
}
\concept{ISSP National Identity}
\keyword{variable}
