% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{shotgun}
\alias{shotgun}
\title{Shotgun in home}
\source{
General Social Survey https://gss.norc.org
}
\description{
shotgun
}
\details{
Question Do you happen to have in your home (IF HOUSE:  or garage) any guns or revolvers?
A. IF YES:  Is it a pistol, shotgun, rifle, or what?  CODE ALL THAT APPLY. (Coded if shotgun)
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
   A/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024 \cr
   Full \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/681/vshow}{GSS Data Explorer page for \code{shotgun}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab refused \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1973 \tab - \tab 413 \tab 1069 \tab 15 \tab 7 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 413 \tab 1056 \tab 10 \tab 5 \tab - \tab - \tab 1484 \cr
   1976 \tab - \tab 418 \tab 1054 \tab 17 \tab 10 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 475 \tab 1042 \tab 2 \tab 11 \tab - \tab - \tab 1530 \cr
   1980 \tab - \tab 432 \tab 1019 \tab 7 \tab 10 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 488 \tab 1340 \tab 26 \tab 6 \tab - \tab - \tab 1860 \cr
   1984 \tab - \tab 407 \tab 1051 \tab 9 \tab 6 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 418 \tab 1102 \tab 10 \tab 4 \tab - \tab - \tab 1534 \cr
   1987 \tab - \tab 474 \tab 1328 \tab 10 \tab 7 \tab - \tab - \tab 1819 \cr
   1988 \tab 504 \tab 232 \tab 728 \tab 10 \tab 7 \tab - \tab - \tab 1481 \cr
   1989 \tab 504 \tab 286 \tab 743 \tab - \tab 4 \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 235 \tab 672 \tab - \tab 9 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 251 \tab 722 \tab - \tab 20 \tab - \tab - \tab 1517 \cr
   1993 \tab 531 \tab 289 \tab 777 \tab 7 \tab 2 \tab - \tab - \tab 1606 \cr
   1994 \tab 996 \tab 480 \tab 1488 \tab 5 \tab 23 \tab - \tab - \tab 2992 \cr
   1996 \tab 981 \tab 473 \tab 1438 \tab 7 \tab 5 \tab - \tab - \tab 2904 \cr
   1998 \tab 951 \tab 389 \tab 1477 \tab 6 \tab 9 \tab - \tab - \tab 2832 \cr
   2000 \tab 956 \tab 343 \tab 1481 \tab 23 \tab 14 \tab - \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 195 \tab 719 \tab 10 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1916 \tab 167 \tab 702 \tab 16 \tab 11 \tab - \tab - \tab 2812 \cr
   2006 \tab 2509 \tab 385 \tab 1578 \tab 35 \tab - \tab 3 \tab - \tab 4510 \cr
   2008 \tab 674 \tab 277 \tab 1049 \tab 19 \tab - \tab 4 \tab - \tab 2023 \cr
   2010 \tab 766 \tab 241 \tab 982 \tab 50 \tab - \tab 5 \tab - \tab 2044 \cr
   2012 \tab 669 \tab 276 \tab 994 \tab 28 \tab - \tab 7 \tab - \tab 1974 \cr
   2014 \tab 827 \tab 293 \tab 1350 \tab 63 \tab - \tab 5 \tab - \tab 2538 \cr
   2016 \tab 978 \tab 346 \tab 1477 \tab 60 \tab - \tab 6 \tab - \tab 2867 \cr
   2018 \tab 779 \tab 319 \tab 1203 \tab 39 \tab 4 \tab 4 \tab - \tab 2348 \cr
   2021 \tab - \tab 732 \tab 3094 \tab 117 \tab 4 \tab 76 \tab 9 \tab 4032 \cr
   2022 \tab 1203 \tab 419 \tab 1813 \tab 72 \tab 6 \tab 24 \tab 7 \tab 3544 \cr
   Total \tab 18564 \tab 10566 \tab 34548 \tab 673 \tab 184 \tab 135 \tab 16 \tab 64686 \cr
}
}

\seealso{
Other Core: 
\code{\link{burglr}},
\code{\link{fear}},
\code{\link{fearhome}},
\code{\link{owngun}},
\code{\link{pistol}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{rifle}},
\code{\link{robbry}}
}
\concept{Core}
\keyword{variable}
