% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{shout}
\alias{shout}
\title{Ppl at work shout at r in hostile manner}
\source{
General Social Survey https://gss.norc.org
}
\description{
shout
}
\details{
Question I. People at work shout or yell at me in a hostile manner.
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3006/vshow}{GSS Data Explorer page for \code{shout}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 31 \tab 122 \tab 241 \tab 1413 \tab 18 \tab 42 \tab 2812 \cr
   2012 \tab 795 \tab 30 \tab 77 \tab 189 \tab 837 \tab 14 \tab 32 \tab 1974 \cr
   Total \tab 1740 \tab 61 \tab 199 \tab 430 \tab 2250 \tab 32 \tab 74 \tab 4786 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
