% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_184.R
\name{showangr}
\alias{showangr}
\title{When i'm angry i let people know.}
\source{
General Social Survey https://gss.norc.org
}
\description{
showangr
}
\details{
Question The following are some statements that people have made. For each one, please tell me if you strongly agree, agree, disagree, or strongly disagree with the statement. (NOTE: INTERVIEWERS SHOULD CODE "neither agree or disagree/it depends" ONLY WHEN VOLUNTEERED.)
E. When I'm angry I let people know.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree or disagree-it depends
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1729/vshow}{GSS Data Explorer page for \code{showangr}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree or disagree-it depends \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 238 \tab 651 \tab 135 \tab 355 \tab 71 \tab 6 \tab 4 \tab 2904 \cr
   Total \tab 1444 \tab 238 \tab 651 \tab 135 \tab 355 \tab 71 \tab 6 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angry}},
\code{\link{badbrks}},
\code{\link{embarrss}},
\code{\link{litcntrl}},
\code{\link{mostluck}},
\code{\link{noemote}},
\code{\link{noplan}},
\code{\link{proud}},
\code{\link{showfeel}}
}
\concept{Emotions}
\keyword{variable}
