% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_651.R
\name{shutbus}
\alias{shutbus}
\title{Govt should shut down business during pandemic}
\source{
General Social Survey https://gss.norc.org
}
\description{
shutbus
}
\details{
Question Do you think the United States government should or should not have the right to do the following at times of severe epidemics? Shut down businesses and places of employment.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should have the right
\item \verb{[2]} probably should have the right
\item \verb{[3]} probably should not have the right
\item \verb{[4]} definitely should not have the right
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7840/vshow}{GSS Data Explorer page for \code{shutbus}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab definitely should have the right \tab probably should have the right \tab probably should not have the right \tab definitely should not have the right \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2382 \tab 367 \tab 310 \tab 164 \tab 214 \tab 96 \tab 2 \tab 9 \tab 3544 \cr
   Total \tab 2382 \tab 367 \tab 310 \tab 164 \tab 214 \tab 96 \tab 2 \tab 9 \tab 3544 \cr
}
}

\keyword{variable}
