% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{sibnum}
\alias{sibnum}
\title{Number of adult sister or brother}
\source{
General Social Survey https://gss.norc.org
}
\description{
sibnum
}
\details{
Question In the first part of this questionnaire, we would like to ask you about your family and friends. For example, about how often you see or visit them, and when you turn to them for help or advice.
We would like to begin with your brothers and sisters. How many adult brothers and/or sisters - we mean brothers or sisters who are age 18 and older - do you have? (We mean brothers and sisters who are still alive. Please include step-brothers and -sisters, half-brothers and sisters and adopted brothers and sisters)
If you have no ADULT brothers or sisters, enter ''0''.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4184/vshow}{GSS Data Explorer page for \code{sibnum}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 12 \tab 13 \tab 15 \tab 16 \tab 17 \tab 19 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 9 \tab not available in this year \tab Total \cr
   2002 \tab 1616 \tab 129 \tab 252 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 259 \tab 185 \tab 96 \tab 85 \tab 48 \tab 74 \tab 6 \tab - \tab 2765 \cr
   Total \tab 1616 \tab 129 \tab 252 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 259 \tab 185 \tab 96 \tab 85 \tab 48 \tab 74 \tab 6 \tab 0 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{bestfrd}},
\code{\link{borrow1}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{comyear}},
\code{\link{cousins}},
\code{\link{cowrkfrd}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{hapunhap}},
\code{\link{helpfrds}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{inlawnum}},
\code{\link{kidcall}},
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidpars}},
\code{\link{kidvisit}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{localgvt}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{malive}},
\code{\link{marcohab}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{neifrd}},
\code{\link{niecenep}},
\code{\link{othfrd}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{palive}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{peocntct}},
\code{\link{posslq}},
\code{\link{posslqy}},
\code{\link{relcall}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibvisit}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{talkedto}},
\code{\link{trustpeo}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{usefrds}},
\code{\link{wantbest}}

Other Siblings: 
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{siborder}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{spsibs}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{Siblings}
\keyword{variable}
