% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{siborder}
\alias{siborder}
\title{Birth order}
\source{
General Social Survey https://gss.norc.org
}
\description{
siborder
}
\details{
Question How many of your (number) brothers and sisters were born before you were?
}
\section{Values}{

\itemize{
\item \verb{[95]} only child
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5082/vshow}{GSS Data Explorer page for \code{siborder}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 16 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab only child \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab 444 \tab 231 \tab 244 \tab 5 \tab 2 \tab 1 \tab 148 \tab 86 \tab 58 \tab 36 \tab 18 \tab 18 \tab 11 \tab 6 \tab 45 \tab 1 \tab 18 \tab - \tab 1372 \cr
   Total \tab 444 \tab 231 \tab 244 \tab 5 \tab 2 \tab 1 \tab 148 \tab 86 \tab 58 \tab 36 \tab 18 \tab 18 \tab 11 \tab 6 \tab 45 \tab 1 \tab 18 \tab 0 \tab 1372 \cr
}
}

\seealso{
Other Siblings: 
\code{\link{sbalive1}},
\code{\link{sbalive2}},
\code{\link{sbalive3}},
\code{\link{sbalive4}},
\code{\link{sbalive5}},
\code{\link{sbalive6}},
\code{\link{sbalive7}},
\code{\link{sbalive8}},
\code{\link{sbalive9}},
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbrel1}},
\code{\link{sbrel2}},
\code{\link{sbrel3}},
\code{\link{sbrel4}},
\code{\link{sbrel5}},
\code{\link{sbrel6}},
\code{\link{sbrel7}},
\code{\link{sbrel8}},
\code{\link{sbrel9}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}},
\code{\link{sbsex7}},
\code{\link{sbsex8}},
\code{\link{sbsex9}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}},
\code{\link{sbyrbrn1}},
\code{\link{sbyrbrn2}},
\code{\link{sbyrbrn3}},
\code{\link{sbyrbrn4}},
\code{\link{sbyrbrn5}},
\code{\link{sbyrbrn6}},
\code{\link{sbyrbrn7}},
\code{\link{sbyrbrn8}},
\code{\link{sbyrbrn9}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{sibnum}},
\code{\link{sibs}},
\code{\link{sibs_7222}},
\code{\link{spsibs}}
}
\concept{Birth Order}
\concept{Siblings}
\keyword{variable}
