% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{singlpar}
\alias{singlpar}
\title{Single parents can raise kids as well as two}
\source{
General Social Survey https://gss.norc.org
}
\description{
singlpar
}
\details{
Question Do you agree or disagree . . .
A. One parent can bring up a child as well as two parents together.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4405/vshow}{GSS Data Explorer page for \code{singlpar}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 124 \tab 387 \tab 203 \tab 542 \tab 159 \tab 10 \tab 22 \tab 2992 \cr
   2002 \tab 1594 \tab 198 \tab 292 \tab 143 \tab 344 \tab 181 \tab 12 \tab 1 \tab 2765 \cr
   2012 \tab 672 \tab 116 \tab 504 \tab 120 \tab 417 \tab 113 \tab 26 \tab 6 \tab 1974 \cr
   Total \tab 3811 \tab 438 \tab 1183 \tab 466 \tab 1303 \tab 453 \tab 48 \tab 29 \tab 7731 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabfst}},
\code{\link{cohabok}},
\code{\link{divbest}},
\code{\link{divifkd1}},
\code{\link{divnokd1}},
\code{\link{fewrksup}},
\code{\link{hubbywk1}},
\code{\link{meovrwrk}},
\code{\link{mrmom}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
