% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_633.R
\name{sinsacri}
\alias{sinsacri}
\title{Jesus' death on cross is only sacrifice that could remove penalty of r's sin}
\source{
General Social Survey https://gss.norc.org
}
\description{
sinsacri
}
\details{
Question Jesus Christ's death on the cross is the only sacrifice that could remove the penalty of my sin.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7532/vshow}{GSS Data Explorer page for \code{sinsacri}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 2258 \tab 692 \tab 356 \tab 240 \tab 97 \tab 97 \tab 6 \tab 162 \tab 26 \tab 98 \tab 4032 \cr
   Total \tab 2258 \tab 692 \tab 356 \tab 240 \tab 97 \tab 97 \tab 6 \tab 162 \tab 26 \tab 98 \tab 4032 \cr
}
}

\keyword{variable}
