% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_451.R
\name{sisnum}
\alias{sisnum}
\title{Number of adult sister's}
\source{
General Social Survey https://gss.norc.org
}
\description{
sisnum
}
\details{
Question How many sisters age 18 or older do you have? (We mean sisters who are still alive; please include step-sisters, half-sisters, and adopted sisters.)
}
\section{Values}{

\itemize{
\item \verb{[0]} none
\item \verb{[1]} one
\item \verb{[2]} two
\item \verb{[3]} three
\item \verb{[4]} four
\item \verb{[5]} five +
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4135/vshow}{GSS Data Explorer page for \code{sisnum}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab none \tab one \tab two \tab three \tab four \tab five + \tab don't know \tab no answer \tab Total \cr
   1986 \tab - \tab 413 \tab 438 \tab 305 \tab 148 \tab 74 \tab 85 \tab 2 \tab 5 \tab 1470 \cr
   Total \tab 0 \tab 413 \tab 438 \tab 305 \tab 148 \tab 74 \tab 85 \tab 2 \tab 5 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{pacall}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{siscall}},
\code{\link{sistime}},
\code{\link{sisvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
