% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_451.R
\name{sistime}
\alias{sistime}
\title{Travel time to sister}
\source{
General Social Survey https://gss.norc.org
}
\description{
sistime
}
\details{
Question How many sisters age 18 or older do you have? (We mean sisters who are still alive; please include step-sisters, half-sisters, and adopted sisters.)
B. About how long would it take you to get to where your sister lives? Think of the time it usually takes door to door.
}
\section{Values}{

\itemize{
\item \verb{[1]} lt 15 min
\item \verb{[2]} 15-30 min
\item \verb{[3]} 30 min-1 hr
\item \verb{[4]} 1-2 hr
\item \verb{[5]} 2-3 hr
\item \verb{[6]} 3-5 hr
\item \verb{[7]} 5-12 hr
\item \verb{[8]} 12+ hr
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4137/vshow}{GSS Data Explorer page for \code{sistime}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab lt 15 min \tab 15-30 min \tab 30 min-1 hr \tab 1-2 hr \tab 2-3 hr \tab 3-5 hr \tab 5-12 hr \tab 12+ hr \tab don't know \tab no answer \tab Total \cr
   1986 \tab 439 \tab 207 \tab 182 \tab 135 \tab 84 \tab 60 \tab 76 \tab 117 \tab 155 \tab 3 \tab 12 \tab 1470 \cr
   Total \tab 439 \tab 207 \tab 182 \tab 135 \tab 84 \tab 60 \tab 76 \tab 117 \tab 155 \tab 3 \tab 12 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{pacall}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sisvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
