% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_451.R
\name{sisvisit}
\alias{sisvisit}
\title{How often visit sister}
\source{
General Social Survey https://gss.norc.org
}
\description{
sisvisit
}
\details{
Question How many sisters age 18 or older do you have? (We mean sisters who are still alive; please include step-sisters, half-sisters, and adopted sisters.)
A. How often do you see or visit with your sister?
}
\section{Values}{

\itemize{
\item \verb{[1]} lives in hshld
\item \verb{[2]} daily
\item \verb{[3]} several x week
\item \verb{[4]} once a week
\item \verb{[5]} once a month
\item \verb{[6]} several x year
\item \verb{[7]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4136/vshow}{GSS Data Explorer page for \code{sisvisit}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab lives in hshld \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab don't know \tab no answer \tab Total \cr
   1986 \tab 413 \tab 25 \tab 67 \tab 95 \tab 145 \tab 214 \tab 229 \tab 272 \tab 1 \tab 9 \tab 1470 \cr
   Total \tab 413 \tab 25 \tab 67 \tab 95 \tab 145 \tab 214 \tab 229 \tab 272 \tab 1 \tab 9 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{pacall}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
