% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{siteyrs}
\alias{siteyrs}
\title{How long have worked for those sites}
\source{
General Social Survey https://gss.norc.org
}
\description{
siteyrs
}
\details{
Question B. How long have you worked full time for those work sites? How many years?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2148/vshow}{GSS Data Explorer page for \code{siteyrs}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 35 \tab 36 \tab 38 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1998 \tab 2220 \tab 68 \tab 45 \tab 27 \tab 17 \tab 17 \tab 12 \tab 14 \tab 8 \tab 8 \tab 6 \tab 7 \tab 5 \tab 60 \tab 14 \tab 2 \tab 6 \tab 2 \tab 4 \tab 9 \tab 1 \tab 1 \tab 4 \tab 1 \tab 47 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 35 \tab 42 \tab 30 \tab 20 \tab 27 \tab 12 \tab 51 \tab 2832 \cr
   Total \tab 2220 \tab 68 \tab 45 \tab 27 \tab 17 \tab 17 \tab 12 \tab 14 \tab 8 \tab 8 \tab 6 \tab 7 \tab 5 \tab 60 \tab 14 \tab 2 \tab 6 \tab 2 \tab 4 \tab 9 \tab 1 \tab 1 \tab 4 \tab 1 \tab 47 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 35 \tab 42 \tab 30 \tab 20 \tab 27 \tab 12 \tab 51 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{wrkoth}}
}
\concept{Jobs}
\keyword{variable}
