% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{skipwork}
\alias{skipwork}
\title{How often r stayed at home or leave work early}
\source{
General Social Survey https://gss.norc.org
}
\description{
skipwork
}
\details{
Question During the past 3 months, how often did you stay at home or leave work early because you were unhappy about your work situation?
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3013/vshow}{GSS Data Explorer page for \code{skipwork}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 14 \tab 89 \tab 231 \tab 1491 \tab 5 \tab 37 \tab 2812 \cr
   Total \tab 945 \tab 14 \tab 89 \tab 231 \tab 1491 \tab 5 \tab 37 \tab 2812 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
