% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{skltrain}
\alias{skltrain}
\title{R had job skill training in thepast 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
skltrain
}
\details{
Question In addition to your main job, do you do any other work for additional income?
IF YES F. Over the past 12 months, have you had any training to improve your job skills?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4625/vshow}{GSS Data Explorer page for \code{skltrain}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 4008 \tab 60 \tab 432 \tab 5 \tab 5 \tab 4510 \cr
   2016 \tab 2533 \tab 51 \tab 259 \tab 1 \tab 23 \tab 2867 \cr
   Total \tab 6541 \tab 111 \tab 691 \tab 6 \tab 28 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{otherwrk}},
\code{\link{spwrkgvt}}
}
\concept{Work Orientation}
\keyword{variable}
