% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{smallbig}
\alias{smallbig}
\title{R preference re: firm size in}
\source{
General Social Survey https://gss.norc.org
}
\description{
smallbig
}
\details{
Question Suppose you were working and could choose between different kinds of jobs. Which of the following would you personally choose?  I would advise...
Please CIRCLE ONE CODE FOR EACH.
B. Working in a small firm (or) working in a large firm
}
\section{Values}{

\itemize{
\item \verb{[1]} small
\item \verb{[2]} big
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4525/vshow}{GSS Data Explorer page for \code{smallbig}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab small \tab big \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 745 \tab 474 \tab 194 \tab 40 \tab 1537 \cr
   1998 \tab 1609 \tab 727 \tab 308 \tab 159 \tab 29 \tab 2832 \cr
   2006 \tab 2992 \tab 1010 \tab 455 \tab 48 \tab 5 \tab 4510 \cr
   Total \tab 4685 \tab 2482 \tab 1237 \tab 401 \tab 74 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
