% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{smnews}
\alias{smnews}
\title{Reliability of news on social media}
\source{
General Social Survey https://gss.norc.org
}
\description{
smnews
}
\details{
Question In general, how reliable would you rate the news on political matters found in the following sources?… Online social media (for example WhatsApp, Facebook, TikTok, Instagram, YouTube, Twitter, Truth Social).
}
\section{Values}{

\itemize{
\item \verb{[1]} very unreliable
\item \verb{[2]} somewhat unreliable
\item \verb{[3]} somewhat reliable
\item \verb{[4]} very reliable
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8165/vshow}{GSS Data Explorer page for \code{smnews}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab very unreliable \tab somewhat unreliable \tab somewhat reliable \tab very reliable \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 404 \tab 486 \tab 491 \tab 55 \tab 187 \tab 1671 \tab 8 \tab 7 \tab 3309 \cr
   Total \tab 0 \tab 404 \tab 486 \tab 491 \tab 55 \tab 187 \tab 1671 \tab 8 \tab 7 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{polnewsfrom}},
\code{\link{radionews}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
