% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{smoke}
\alias{smoke}
\title{Does r smoke}
\source{
General Social Survey https://gss.norc.org
}
\description{
smoke
}
\details{
Question Do you smoke?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1977, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/532/vshow}{GSS Data Explorer page for \code{smoke}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1977 \tab - \tab 646 \tab 878 \tab 6 \tab 1530 \cr
   1978 \tab - \tab 604 \tab 927 \tab 1 \tab 1532 \cr
   1980 \tab - \tab 601 \tab 866 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 591 \tab 1006 \tab 2 \tab 1599 \cr
   1984 \tab - \tab 542 \tab 925 \tab 6 \tab 1473 \cr
   1986 \tab - \tab 477 \tab 989 \tab 4 \tab 1470 \cr
   1987 \tab - \tab 582 \tab 1232 \tab 5 \tab 1819 \cr
   1988 \tab 484 \tab 339 \tab 656 \tab 2 \tab 1481 \cr
   1989 \tab 502 \tab 309 \tab 724 \tab 2 \tab 1537 \cr
   1990 \tab 473 \tab 285 \tab 612 \tab 2 \tab 1372 \cr
   1991 \tab 500 \tab 319 \tab 697 \tab 1 \tab 1517 \cr
   1993 \tab 549 \tab 288 \tab 767 \tab 2 \tab 1606 \cr
   1994 \tab 2481 \tab 142 \tab 369 \tab - \tab 2992 \cr
   Total \tab 4989 \tab 5725 \tab 10648 \tab 34 \tab 21396 \cr
}
}

\seealso{
Other Core: 
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{partfull}},
\code{\link{socpars}},
\code{\link{socsibs}},
\code{\link{weekswrk}}
}
\concept{Core}
\keyword{variable}
