% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_286.R
\name{snapchat}
\alias{snapchat}
\title{Use snapchat}
\source{
General Social Survey https://gss.norc.org
}
\description{
snapchat
}
\details{
Question (Which of the following social networking or social media sites are you a member or regular user of?)
Snapchat.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6273/vshow}{GSS Data Explorer page for \code{snapchat}}.
}

\seealso{
Other Social Media Usage: 
\code{\link{facebook}},
\code{\link{googlesn}},
\code{\link{instagrm}},
\code{\link{intrecnt}},
\code{\link{linkedin}},
\code{\link{tumblr}},
\code{\link{twitter}},
\code{\link{whatsapp}}
}
\concept{Social Media Usage}
\keyword{variable}
