% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_287.R
\name{snsmot2c}
\alias{snsmot2c}
\title{Third other social network}
\source{
General Social Survey https://gss.norc.org
}
\description{
snsmot2c
}
\details{
Question THIRD MENTION:
}
\section{Values}{

\itemize{
\item \verb{[1]} youtube
\item \verb{[2]} myspace
\item \verb{[3]} reddit
\item \verb{[4]} viber
\item \verb{[5]} line
\item \verb{[6]} plenty of fish
\item \verb{[7]} kik
\item \verb{[8]} ravelry
\item \verb{[9]} wechat
\item \verb{[11]} imo
\item \verb{[12]} online dating sites
\item \verb{[13]} spotify
\item \verb{[14]} groupme
\item \verb{[15]} tinder
\item \verb{[16]} hangouts
\item \verb{[17]} researchgate
\item \verb{[18]} online gaming
\item \verb{[19]} yahoo
\item \verb{[30]} other social networks
\item \verb{[40]} not a social network
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6284/vshow}{GSS Data Explorer page for \code{snsmot2c}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab youtube \tab viber \tab kik \tab hangouts \tab other social networks \tab not a social network \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2845 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 11 \tab 2867 \cr
   Total \tab 2845 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 11 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{clssmtes}},
\code{\link{flickr}},
\code{\link{pinterst}},
\code{\link{snsmnew}},
\code{\link{snsmot2a}},
\code{\link{snsmot2b}},
\code{\link{snsmoth1}},
\code{\link{vine}}
}
\concept{Social Media Usage}
\keyword{variable}
