% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{socacts}
\alias{socacts}
\title{Physical and emotion social activities past 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
socacts
}
\details{
Question During the past 4 weeks, how much of the time has your physical health or emotional problems interfered with your social activities (like visiting friends, relatives, etc.)?
}
\section{Values}{

\itemize{
\item \code{1} all of the time
\item \code{2} most of the time
\item \code{3} a good bit of the time
\item \code{4} some of the time
\item \code{5} a little bit of the time
\item \code{6} non of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2687/vshow}{GSS Data Explorer page for \code{socacts}}.
}

\seealso{
Other Health Status: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{pain}},
\code{\link{peaceful}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}

Other Health: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{pain}},
\code{\link{peaceful}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}
}
\concept{Health}
\concept{Health Status}
\keyword{variable}
