% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{socdif1}
\alias{socdif1}
\title{Income differences an incentive}
\source{
General Social Survey https://gss.norc.org
}
\description{
socdif1
}
\details{
Question Here are different opinions about social differences in this country. Please tell me for each one whether you strongly agree, somewhat agree, somewhat disagree, or strongly disagree.
A. Only if differences in income are large enough is there an incentive for individual effort
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} disagree
\item \code{4} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1174/vshow}{GSS Data Explorer page for \code{socdif1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab disagree \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 79 \tab 632 \tab 454 \tab 61 \tab 141 \tab 5 \tab 1372 \cr
   Total \tab 0 \tab 79 \tab 632 \tab 454 \tab 61 \tab 141 \tab 5 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthjews}},
\code{\link{wlthwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
