% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{socid2}
\alias{socid2}
\title{Second most important to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
socid2
}
\details{
Question We are all part of different groups. Some are more important to us than others when we think of ourselves. in general, which in the following list is most important to you in describing who you are?  and the second most important? and the third most important?
B. Second Most Important
}
\section{Values}{

\itemize{
\item \code{1} current occupation
\item \code{2} race ethnic background
\item \code{3} gender
\item \code{4} age group
\item \code{5} religion
\item \code{6} political party
\item \code{7} nationality
\item \code{8} family or marital status
\item \code{9} social class
\item \code{10} region
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4879/vshow}{GSS Data Explorer page for \code{socid2}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab current occupation \tab race ethnic background \tab gender \tab age group \tab religion \tab political party \tab nationality \tab family or marital status \tab social class \tab region \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 236 \tab 74 \tab 102 \tab 118 \tab 158 \tab 14 \tab 63 \tab 236 \tab 75 \tab 120 \tab 13 \tab 7 \tab 2812 \cr
   Total \tab 1596 \tab 236 \tab 74 \tab 102 \tab 118 \tab 158 \tab 14 \tab 63 \tab 236 \tab 75 \tab 120 \tab 13 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
