% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{socright}
\alias{socright}
\title{Society's laws should decide right and wrong}
\source{
General Social Survey https://gss.norc.org
}
\description{
socright
}
\details{
Question Do you agree or disagree with the following statements. . .
B. Right and wrong should be decided by society.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4682/vshow}{GSS Data Explorer page for \code{socright}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 113 \tab 418 \tab 309 \tab 241 \tab 142 \tab 59 \tab 77 \tab 1517 \cr
   Total \tab 158 \tab 113 \tab 418 \tab 309 \tab 241 \tab 142 \tab 59 \tab 77 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
