% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{socsecrt}
\alias{socsecrt}
\title{Social security will be better}
\source{
General Social Survey https://gss.norc.org
}
\description{
socsecrt
}
\details{
Question When you retire, do you think Social Security benefits will be much better than those now paid of retirees, somewhat better, about the same, somewhat worse, or much worse?
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} somewhat better
\item \code{3} about the same
\item \code{4} somewhat worse
\item \code{5} much worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2242/vshow}{GSS Data Explorer page for \code{socsecrt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much better \tab somewhat better \tab about the same \tab somewhat worse \tab much worse \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 32 \tab 86 \tab 318 \tab 360 \tab 420 \tab 100 \tab 71 \tab 2832 \cr
   Total \tab 1445 \tab 32 \tab 86 \tab 318 \tab 360 \tab 420 \tab 100 \tab 71 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
