% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{solok}
\alias{solok}
\title{How important: citizens have adequate standard of living}
\source{
General Social Survey https://gss.norc.org
}
\description{
solok
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
A. That all citizens have an adequate standard of living
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4933/vshow}{GSS Data Explorer page for \code{solok}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2004 \tab 1340 \tab 10 \tab 9 \tab 22 \tab 83 \tab 184 \tab 200 \tab 954 \tab 5 \tab 5 \tab - \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 13 \tab 7 \tab 24 \tab 95 \tab 179 \tab 201 \tab 714 \tab 29 \tab 2 \tab - \tab - \tab 2538 \cr
   2024 \tab 1663 \tab 24 \tab 26 \tab 48 \tab 121 \tab 221 \tab 268 \tab 846 \tab 70 \tab 3 \tab - \tab 19 \tab 3309 \cr
   Total \tab 4277 \tab 47 \tab 42 \tab 94 \tab 299 \tab 584 \tab 669 \tab 2514 \tab 104 \tab 10 \tab 0 \tab 19 \tab 8659 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{polopts}},
\code{\link{rghtsmin}}

Other Democracy: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{polopts}},
\code{\link{rghtsmin}}
}
\concept{Democracy}
\concept{ISSP Citizenship}
\keyword{variable}
