% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_108.R
\name{solprof}
\alias{solprof}
\title{Prof society tries to solve problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
solprof
}
\details{
Question Do (EACH "YES" IN  MEMFRAT-MEMNUM) to which you belong do anything to try to solve individual or community problems?
N. Professional or academic societies
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/993/vshow}{GSS Data Explorer page for \code{solprof}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{actfrat}},
\code{\link{actpolit}},
\code{\link{actserv}},
\code{\link{actvet}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}}
}
\concept{Socio-political participation}
\keyword{variable}
