% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_333.R
\name{solveown}
\alias{solveown}
\title{R would solve problem on ones own or get advice}
\source{
General Social Survey https://gss.norc.org
}
\description{
solveown
}
\details{
Question When you have a problem, do you prefer to solve it on your own or do you like to talk to other people to get advice, help or information?
}
\section{Values}{

\itemize{
\item \verb{[1]} on own
\item \verb{[2]} get advice
\item \verb{[3]} both(on own and get advice) depending on problem
\item \verb{[4]} religious
\item \verb{[5]} research the problem
\item \verb{[6]} talk to a doctor other professional
\item \verb{[7]} talk with older people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2978/vshow}{GSS Data Explorer page for \code{solveown}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab on own \tab get advice \tab both(on own and get advice) depending on problem \tab religious \tab research the problem \tab talk to a doctor other professional \tab talk with older people \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2002 \tab 1372 \tab 593 \tab 729 \tab 35 \tab 7 \tab 3 \tab 4 \tab 1 \tab 14 \tab 7 \tab - \tab 2765 \cr
   Total \tab 1372 \tab 593 \tab 729 \tab 35 \tab 7 \tab 3 \tab 4 \tab 1 \tab 14 \tab 7 \tab 0 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{allergic}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{chembal}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldprob}},
\code{\link{chldsch}},
\code{\link{chldvig}},
\code{\link{comknows}},
\code{\link{discipln}},
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{failure}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{illphys}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{medaddct}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mntlill}},
\code{\link{mntlrel}},
\code{\link{nextdoor}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{pryntfam}},
\code{\link{putsoff}},
\code{\link{raised}},
\code{\link{spendeve}},
\code{\link{stressfl}},
\code{\link{sufadult}},
\code{\link{trbllaw}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{zombies}}

Other Life Events: 
\code{\link{attrelig}},
\code{\link{eatout}},
\code{\link{extra1}},
\code{\link{extra2}},
\code{\link{famper6}},
\code{\link{godoc}},
\code{\link{prob1}},
\code{\link{prob2}},
\code{\link{prob3}},
\code{\link{prob4}},
\code{\link{seefilm}}
}
\concept{Life Events}
\concept{Mental Health II}
\keyword{variable}
