% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_107.R
\name{solyouth}
\alias{solyouth}
\title{Youth group tries to solve problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
solyouth
}
\details{
Question Do (EACH "YES" IN  MEMFRAT-MEMNUM) to which you belong do anything to try to solve individual or community problems?
G. Youth groups
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/986/vshow}{GSS Data Explorer page for \code{solyouth}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab 1645 \tab 114 \tab 30 \tab 30 \tab 1819 \cr
   Total \tab 1645 \tab 114 \tab 30 \tab 30 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{solpolit}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}}
}
\concept{Socio-political participation}
\keyword{variable}
