% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{soncall}
\alias{soncall}
\title{Contact with son}
\source{
General Social Survey https://gss.norc.org
}
\description{
soncall
}
\details{
Question How many sons age 18 or older do you have? (We mean sons who are still alive; please include step-sons, half-sons, and adopted sons.)
C. And how often do you have any other contact with your son besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several x week
\item \verb{[3]} once a week
\item \verb{[4]} once a month
\item \verb{[5]} several x year
\item \verb{[6]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4150/vshow}{GSS Data Explorer page for \code{soncall}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab no answer \tab Total \cr
   1986 \tab 1076 \tab 57 \tab 86 \tab 100 \tab 73 \tab 37 \tab 34 \tab 7 \tab 1470 \cr
   Total \tab 1076 \tab 57 \tab 86 \tab 100 \tab 73 \tab 37 \tab 34 \tab 7 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{brocall}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{posslq}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
