% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{sonvisit}
\alias{sonvisit}
\title{How often visit son}
\source{
General Social Survey https://gss.norc.org
}
\description{
sonvisit
}
\details{
Question How many sons age 18 or older do you have? (We mean sons who are still alive; please include step-sons, half-sons, and adopted sons.)
A. How often do you see or visit with your son?
}
\section{Values}{

\itemize{
\item \verb{[1]} lives in hshld
\item \verb{[2]} daily
\item \verb{[3]} several x week
\item \verb{[4]} once a week
\item \verb{[5]} once a month
\item \verb{[6]} several x year
\item \verb{[7]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4148/vshow}{GSS Data Explorer page for \code{sonvisit}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab lives in hshld \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab no answer \tab Total \cr
   1986 \tab 973 \tab 103 \tab 59 \tab 57 \tab 70 \tab 65 \tab 79 \tab 59 \tab 5 \tab 1470 \cr
   Total \tab 973 \tab 103 \tab 59 \tab 57 \tab 70 \tab 65 \tab 79 \tab 59 \tab 5 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{brocall}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{posslq}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
