% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{spaneng}
\alias{spaneng}
\title{Interviews conducted in spanish or english}
\source{
General Social Survey https://gss.norc.org
}
\description{
spaneng
}
\details{
Question Interviews Conducted in Spanish or English
}
\section{Values}{

\itemize{
\item \verb{[1]} english
\item \verb{[2]} spanish
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5584/vshow}{GSS Data Explorer page for \code{spaneng}}.

Counts by year:\tabular{lllll}{
   year \tab english \tab spanish \tab no answer \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab 2812 \cr
   2006 \tab 4233 \tab 277 \tab - \tab 4510 \cr
   2008 \tab 1942 \tab 81 \tab - \tab 2023 \cr
   2010 \tab 1971 \tab 73 \tab - \tab 2044 \cr
   2012 \tab 1892 \tab 82 \tab - \tab 1974 \cr
   2014 \tab 2423 \tab 109 \tab 6 \tab 2538 \cr
   2016 \tab 2773 \tab 94 \tab - \tab 2867 \cr
   2018 \tab 2258 \tab 90 \tab - \tab 2348 \cr
   2021 \tab 3976 \tab 56 \tab - \tab 4032 \cr
   2022 \tab 3431 \tab 113 \tab - \tab 3544 \cr
   Total \tab 71409 \tab 975 \tab 6 \tab 72390 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
