% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{spanint}
\alias{spanint}
\title{If no spanish, r could have been interviewed in english}
\source{
General Social Survey https://gss.norc.org
}
\description{
spanint
}
\details{
Question Before 2006 the GSS was only administered in English. Those without enough English to do the interview were excluded as out-of-scope, language problems. In 2006 a Spanish version of the GSS was added. If there had been no Spanish version available in 2006, could this respondent have been interviewed in English or do you think (s/he) would have been excluded as a language problem?
}
\section{Values}{

\itemize{
\item \code{1} could have been interviewed in english
\item \code{2} would have been excluded as language problem
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab B/C/-/- \tab partial \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5583/vshow}{GSS Data Explorer page for \code{spanint}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{batch}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{cshutyp12}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{spaneng}},
\code{\link{spanself}},
\code{\link{subsamprate}},
\code{\link{version}},
\code{\link{vpsu}},
\code{\link{vstrat}},
\code{\link{wtss}},
\code{\link{wtssall}},
\code{\link{wtssnr}}
}
\concept{Sampling and Weights}
\keyword{variable}
