% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{spanself}
\alias{spanself}
\title{If this interview had only been available in english, would you...}
\source{
General Social Survey https://gss.norc.org
}
\description{
spanself
}
\details{
Question If this interview had only been available in English, would you…
}
\section{Values}{

\itemize{
\item \verb{[1]} have been able to do the interview easily in english
\item \verb{[2]} have been able to do the interview with difficulty
\item \verb{[3]} not have been able to do the interview
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5582/vshow}{GSS Data Explorer page for \code{spanself}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab have been able to do the interview easily in english \tab have been able to do the interview with difficulty \tab not have been able to do the interview \tab no answer \tab Total \cr
   2006 \tab 4233 \tab 28 \tab 57 \tab 192 \tab - \tab 4510 \cr
   2008 \tab 1943 \tab 7 \tab 18 \tab 54 \tab 1 \tab 2023 \cr
   2010 \tab 1971 \tab 6 \tab 17 \tab 50 \tab - \tab 2044 \cr
   2012 \tab 1894 \tab 4 \tab 16 \tab 59 \tab 1 \tab 1974 \cr
   2014 \tab 2427 \tab 15 \tab 22 \tab 74 \tab - \tab 2538 \cr
   2016 \tab 2775 \tab 5 \tab 23 \tab 64 \tab - \tab 2867 \cr
   2018 \tab 2260 \tab 7 \tab 21 \tab 60 \tab - \tab 2348 \cr
   2021 \tab 4005 \tab 5 \tab 15 \tab 7 \tab - \tab 4032 \cr
   2022 \tab 3433 \tab 19 \tab 30 \tab 62 \tab - \tab 3544 \cr
   Total \tab 24941 \tab 96 \tab 219 \tab 622 \tab 2 \tab 25880 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{spaneng}},
\code{\link{spanint}}
}
\concept{Sampling and Weights}
\keyword{variable}
