% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{spattend}
\alias{spattend}
\title{How often sp attends religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
spattend
}
\details{
Question C. How often does your (husband/wife) attend religious services?
(USE CATEGORIES AS PROBES, IF NECESSARY.)
}
\section{Values}{

\itemize{
\item \code{0} never
\item \code{1} lt once a year
\item \code{2} once a year
\item \code{3} sevrl times a yr
\item \code{4} once a month
\item \code{5} 2-3x a month
\item \code{6} nrly every week
\item \code{7} every week
\item \code{8} more thn once wk
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/297/vshow}{GSS Data Explorer page for \code{spattend}}.
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jew}},
\code{\link{jewaj}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{relitena}}
}
\concept{Core}
\keyword{variable}
