% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{spbarate}
\alias{spbarate}
\title{Rate of ba degree earning in college spouse attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
spbarate
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
K. Graduation Rate
}
\section{Values}{

\itemize{
\item \verb{[1]} 0-39.9pct
\item \verb{[2]} 40-57.9pct
\item \verb{[3]} 58pct or over
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/79/vshow}{GSS Data Explorer page for \code{spbarate}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 0-39.9pct \tab 40-57.9pct \tab 58pct or over \tab no answer \tab Total \cr
   2012 \tab 1744 \tab 30 \tab 80 \tab 119 \tab 1 \tab 1974 \cr
   2014 \tab 2168 \tab 63 \tab 123 \tab 183 \tab 1 \tab 2538 \cr
   2016 \tab 2511 \tab 45 \tab 101 \tab 210 \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 8771 \tab 138 \tab 304 \tab 512 \tab 2 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
