% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{spdays}
\alias{spdays}
\title{Days spouse normally works}
\source{
General Social Survey https://gss.norc.org
}
\description{
spdays
}
\details{
Question What days of the week does (SPOUSE) normally work?  CODE ALL THAT APPLY.
}
\section{Values}{

\itemize{
\item \verb{[1]} mon to fri
\item \verb{[2]} mon to fri,sat
\item \verb{[3]} mon to fri,sun
\item \verb{[4]} wkend plus
\item \verb{[5]} no schedule
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5289/vshow}{GSS Data Explorer page for \code{spdays}}.

Counts by year:\tabular{lllllllll}{
   year \tab mon to fri \tab mon to fri,sat \tab mon to fri,sun \tab wkend plus \tab no schedule \tab iap \tab no answer \tab Total \cr
   1972 \tab 345 \tab 116 \tab 7 \tab 45 \tab 52 \tab 1044 \tab 4 \tab 1613 \cr
   Total \tab 345 \tab 116 \tab 7 \tab 45 \tab 52 \tab 1044 \tab 4 \tab 1613 \cr
}
}

\seealso{
Other Military Service & Work: 
\code{\link{sphour}},
\code{\link{vetkind}},
\code{\link{vetyears}},
\code{\link{workdy}},
\code{\link{workhr}}
}
\concept{Military Service & Work}
\keyword{variable}
