% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_84.R
\name{spdeath}
\alias{spdeath}
\title{Death of spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
spdeath
}
\details{
Question E. Death of spouse
}
\section{Values}{

\itemize{
\item \verb{[0]} no death reported
\item \verb{[1]} died 6+ yrs ago
\item \verb{[2]} died 1-5 yrs ago
\item \verb{[3]} died last year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/752/vshow}{GSS Data Explorer page for \code{spdeath}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no death reported \tab died 6+ yrs ago \tab died 1-5 yrs ago \tab died last year \tab no answer \tab Total \cr
   1978 \tab 237 \tab 1102 \tab 115 \tab 41 \tab 18 \tab 19 \tab 1532 \cr
   1980 \tab 242 \tab 1037 \tab 109 \tab 45 \tab 22 \tab 13 \tab 1468 \cr
   1983 \tab 265 \tab 1135 \tab 123 \tab 41 \tab 15 \tab 20 \tab 1599 \cr
   1984 \tab 282 \tab 1001 \tab 101 \tab 54 \tab 17 \tab 18 \tab 1473 \cr
   1986 \tab 266 \tab 1009 \tab 129 \tab 44 \tab 18 \tab 4 \tab 1470 \cr
   1987 \tab 380 \tab 1179 \tab 152 \tab 54 \tab 36 \tab 18 \tab 1819 \cr
   1988 \tab 675 \tab 682 \tab 80 \tab 20 \tab 17 \tab 7 \tab 1481 \cr
   1989 \tab 710 \tab 704 \tab 67 \tab 33 \tab 16 \tab 7 \tab 1537 \cr
   1990 \tab 650 \tab 596 \tab 86 \tab 20 \tab 13 \tab 7 \tab 1372 \cr
   1991 \tab 710 \tab 673 \tab 79 \tab 39 \tab 6 \tab 10 \tab 1517 \cr
   1993 \tab 736 \tab 726 \tab 86 \tab 31 \tab 13 \tab 14 \tab 1606 \cr
   1994 \tab 2585 \tab 350 \tab 36 \tab 14 \tab 2 \tab 5 \tab 2992 \cr
   Total \tab 7738 \tab 10194 \tab 1163 \tab 436 \tab 193 \tab 142 \tab 19866 \cr
}
}

\seealso{
Other Core: 
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{sibdeath}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}
}
\concept{Core}
\keyword{variable}
