% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{spdeg}
\alias{spdeg}
\title{Spouse's highest degree}
\source{
General Social Survey https://gss.norc.org
}
\description{
spdeg
}
\details{
Question RESPONDENT'S SPOUSE'S DEGREE
}
\section{Values}{

\itemize{
\item \code{0} less than high school
\item \code{1} high school
\item \code{2} associate/junior college
\item \code{3} bachelor's
\item \code{4} graduate
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/62/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab associate/junior college \tab bachelor's \tab don't know \tab graduate \tab high school \tab iap \tab less than high school \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 10 \tab 92 \tab 15 \tab 37 \tab 568 \tab 453 \tab 403 \tab 35 \tab - \tab 1613 \cr
   1973 \tab 20 \tab 100 \tab 2 \tab 53 \tab 500 \tab 428 \tab 379 \tab 22 \tab - \tab 1504 \cr
   1974 \tab 20 \tab 95 \tab 5 \tab 42 \tab 526 \tab 419 \tab 373 \tab 4 \tab - \tab 1484 \cr
   1975 \tab 17 \tab 92 \tab 3 \tab 50 \tab 524 \tab 488 \tab 315 \tab 1 \tab - \tab 1490 \cr
   1976 \tab 21 \tab 79 \tab 2 \tab 50 \tab 483 \tab 525 \tab 336 \tab 3 \tab - \tab 1499 \cr
   1977 \tab 27 \tab 97 \tab 4 \tab 45 \tab 478 \tab 555 \tab 319 \tab 5 \tab - \tab 1530 \cr
   1978 \tab 29 \tab 90 \tab 7 \tab 55 \tab 508 \tab 571 \tab 268 \tab 4 \tab - \tab 1532 \cr
   1980 \tab 24 \tab 82 \tab 2 \tab 56 \tab 469 \tab 579 \tab 250 \tab 6 \tab - \tab 1468 \cr
   1982 \tab 24 \tab 115 \tab 8 \tab 50 \tab 521 \tab 867 \tab 272 \tab 3 \tab - \tab 1860 \cr
   1983 \tab 34 \tab 117 \tab 4 \tab 58 \tab 535 \tab 634 \tab 216 \tab 1 \tab - \tab 1599 \cr
   1984 \tab 30 \tab 95 \tab 2 \tab 48 \tab 459 \tab 644 \tab 194 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 48 \tab 99 \tab 3 \tab 49 \tab 469 \tab 663 \tab 202 \tab 1 \tab - \tab 1534 \cr
   1986 \tab 38 \tab 114 \tab 3 \tab 48 \tab 440 \tab 644 \tab 183 \tab - \tab - \tab 1470 \cr
   1987 \tab 35 \tab 123 \tab 9 \tab 52 \tab 501 \tab 903 \tab 191 \tab 5 \tab - \tab 1819 \cr
   1988 \tab 30 \tab 85 \tab 3 \tab 57 \tab 463 \tab 693 \tab 150 \tab - \tab - \tab 1481 \cr
   1989 \tab 43 \tab 115 \tab 4 \tab 53 \tab 457 \tab 690 \tab 171 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 32 \tab 107 \tab 2 \tab 56 \tab 390 \tab 645 \tab 135 \tab 5 \tab - \tab 1372 \cr
   1991 \tab 33 \tab 128 \tab 6 \tab 44 \tab 436 \tab 713 \tab 153 \tab 4 \tab - \tab 1517 \cr
   1993 \tab 44 \tab 148 \tab 5 \tab 72 \tab 454 \tab 747 \tab 133 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 102 \tab 278 \tab 5 \tab 115 \tab 817 \tab 1451 \tab 214 \tab 10 \tab - \tab 2992 \cr
   1996 \tab 99 \tab 218 \tab 2 \tab 126 \tab 753 \tab 1514 \tab 189 \tab 3 \tab - \tab 2904 \cr
   1998 \tab 103 \tab 237 \tab 12 \tab 121 \tab 693 \tab 1487 \tab 173 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 95 \tab 209 \tab 9 \tab 116 \tab 665 \tab 1538 \tab 179 \tab 6 \tab - \tab 2817 \cr
   2002 \tab 83 \tab 241 \tab 8 \tab 135 \tab 653 \tab 1496 \tab 149 \tab - \tab - \tab 2765 \cr
   2004 \tab 115 \tab 292 \tab 5 \tab 173 \tab 733 \tab 1333 \tab 156 \tab 5 \tab - \tab 2812 \cr
   2006 \tab 121 \tab 274 \tab 7 \tab 147 \tab 640 \tab 3095 \tab 219 \tab 7 \tab - \tab 4510 \cr
   2008 \tab 71 \tab 195 \tab 5 \tab 105 \tab 499 \tab 1046 \tab 97 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 68 \tab 177 \tab 2 \tab 105 \tab 435 \tab 1152 \tab 103 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 56 \tab 169 \tab 2 \tab 112 \tab 444 \tab 1074 \tab 117 \tab - \tab - \tab 1974 \cr
   2014 \tab 89 \tab 238 \tab 4 \tab 169 \tab 529 \tab 1376 \tab 129 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 89 \tab 261 \tab 6 \tab 163 \tab 578 \tab 1654 \tab 113 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 79 \tab 221 \tab 2 \tab 138 \tab 443 \tab 1348 \tab 112 \tab 5 \tab - \tab 2348 \cr
   2021 \tab 146 \tab 517 \tab 19 \tab 388 \tab 769 \tab 2029 \tab 151 \tab - \tab 13 \tab 4032 \cr
   2022 \tab 106 \tab 339 \tab 21 \tab 273 \tab 618 \tab 2074 \tab 110 \tab 2 \tab 1 \tab 3544 \cr
   2024 \tab 124 \tab 345 \tab 4 \tab 241 \tab 527 \tab 1957 \tab 111 \tab - \tab - \tab 3309 \cr
   Total \tab 2105 \tab 6184 \tab 202 \tab 3602 \tab 18977 \tab 37485 \tab 6965 \tab 165 \tab 14 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{codipged}},
\code{\link{cosector}},
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}

Other Education: 
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\concept{Education}
\keyword{variable}
