% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{spden}
\alias{spden}
\title{Specific denomination, spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
spden
}
\details{
Question What is your (SPOUSE''S) religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?
A. What specific denomination is that, if any? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \verb{[1]} baptist
\item \verb{[2]} methodist
\item \verb{[3]} lutheran
\item \verb{[4]} presbyterian
\item \verb{[5]} episcopal
\item \verb{[6]} other denomination
\item \verb{[7]} no denomination or non-denominational church
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   Full \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/349/vshow}{GSS Data Explorer page for \code{spden}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab baptist \tab methodist \tab lutheran \tab presbyterian \tab episcopal \tab other denomination \tab no denomination or non-denominational church \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1973 \tab - \tab 207 \tab 165 \tab 79 \tab 49 \tab 33 \tab 142 \tab 32 \tab 796 \tab 1 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 219 \tab 148 \tab 94 \tab 62 \tab 27 \tab 120 \tab 31 \tab 779 \tab 4 \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 214 \tab 108 \tab 103 \tab 59 \tab 28 \tab 117 \tab 45 \tab 814 \tab 1 \tab 1 \tab - \tab 1490 \cr
   1976 \tab - \tab 190 \tab 123 \tab 82 \tab 47 \tab 29 \tab 125 \tab 33 \tab 864 \tab 4 \tab 2 \tab - \tab 1499 \cr
   1977 \tab - \tab 206 \tab 124 \tab 83 \tab 52 \tab 20 \tab 117 \tab 37 \tab 888 \tab 3 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 190 \tab 110 \tab 81 \tab 45 \tab 20 \tab 126 \tab 41 \tab 914 \tab 4 \tab 1 \tab - \tab 1532 \cr
   1980 \tab - \tab 177 \tab 105 \tab 51 \tab 53 \tab 24 \tab 125 \tab 25 \tab 897 \tab 4 \tab 7 \tab - \tab 1468 \cr
   1982 \tab - \tab 236 \tab 113 \tab 78 \tab 41 \tab 25 \tab 134 \tab 32 \tab 1193 \tab 8 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 168 \tab 89 \tab 89 \tab 51 \tab 22 \tab 149 \tab 32 \tab 993 \tab 6 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 147 \tab 88 \tab 68 \tab 42 \tab 15 \tab 141 \tab 35 \tab 926 \tab 11 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 173 \tab 95 \tab 72 \tab 41 \tab 24 \tab 124 \tab 30 \tab 970 \tab 5 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 153 \tab 85 \tab 63 \tab 37 \tab 23 \tab 116 \tab 35 \tab 953 \tab 4 \tab 1 \tab - \tab 1470 \cr
   1987 \tab - \tab 225 \tab 108 \tab 53 \tab 43 \tab 14 \tab 149 \tab 31 \tab 1185 \tab 11 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 171 \tab 78 \tab 51 \tab 35 \tab 16 \tab 129 \tab 35 \tab 966 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 160 \tab 81 \tab 61 \tab 32 \tab 21 \tab 107 \tab 52 \tab 1013 \tab 10 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 106 \tab 52 \tab 64 \tab 40 \tab 16 \tab 120 \tab 43 \tab 925 \tab 5 \tab 1 \tab - \tab 1372 \cr
   1991 \tab - \tab 150 \tab 76 \tab 47 \tab 41 \tab 19 \tab 137 \tab 27 \tab 1016 \tab 4 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 155 \tab 95 \tab 67 \tab 31 \tab 18 \tab 123 \tab 46 \tab 1061 \tab 9 \tab 1 \tab - \tab 1606 \cr
   1994 \tab - \tab 142 \tab 82 \tab 57 \tab 36 \tab 22 \tab 87 \tab 34 \tab 2506 \tab 23 \tab 3 \tab - \tab 2992 \cr
   2004 \tab - \tab 119 \tab 46 \tab 35 \tab 19 \tab 8 \tab 80 \tab 68 \tab 2431 \tab 2 \tab 4 \tab - \tab 2812 \cr
   2006 \tab - \tab 227 \tab 90 \tab 60 \tab 42 \tab 23 \tab 195 \tab 95 \tab 3765 \tab 2 \tab 11 \tab - \tab 4510 \cr
   2008 \tab - \tab 151 \tab 63 \tab 44 \tab 29 \tab 27 \tab 140 \tab 95 \tab 1461 \tab 10 \tab 3 \tab - \tab 2023 \cr
   2010 \tab - \tab 151 \tab 47 \tab 39 \tab 29 \tab 26 \tab 99 \tab 72 \tab 1573 \tab 1 \tab 7 \tab - \tab 2044 \cr
   2012 \tab - \tab 127 \tab 54 \tab 41 \tab 23 \tab 15 \tab 125 \tab 104 \tab 1476 \tab 2 \tab 7 \tab - \tab 1974 \cr
   2014 \tab - \tab 136 \tab 79 \tab 56 \tab 28 \tab 17 \tab 115 \tab 152 \tab 1936 \tab 12 \tab 7 \tab - \tab 2538 \cr
   2016 \tab - \tab 157 \tab 56 \tab 58 \tab 19 \tab 12 \tab 137 \tab 130 \tab 2279 \tab 7 \tab 12 \tab - \tab 2867 \cr
   2018 \tab - \tab 133 \tab 53 \tab 32 \tab 25 \tab 11 \tab 112 \tab 114 \tab 1853 \tab 4 \tab 11 \tab - \tab 2348 \cr
   2021 \tab - \tab 190 \tab 95 \tab 92 \tab 53 \tab 27 \tab 123 \tab 250 \tab 3181 \tab 3 \tab 15 \tab 3 \tab 4032 \cr
   2022 \tab - \tab 165 \tab 60 \tab 51 \tab 34 \tab 19 \tab 97 \tab 193 \tab 2902 \tab 23 \tab - \tab - \tab 3544 \cr
   Total \tab 0 \tab 4945 \tab 2568 \tab 1851 \tab 1138 \tab 601 \tab 3611 \tab 1949 \tab 42516 \tab 183 \tab 94 \tab 3 \tab 59459 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom16}},
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spother}},
\code{\link{sprel}}

Other Religion: 
\code{\link{denom16}},
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spother}},
\code{\link{sprel}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
