% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{spdue}
\alias{spdue}
\title{Ever entitled to alimony or chld support}
\source{
General Social Survey https://gss.norc.org
}
\description{
spdue
}
\details{
Question Are you currently or have you ever been entitled to alimony or child support payments from a former (husband/wife)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/629/vshow}{GSS Data Explorer page for \code{spdue}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1986 \tab 1072 \tab 124 \tab 244 \tab 30 \tab 1470 \cr
   Total \tab 1072 \tab 124 \tab 244 \tab 30 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{sppaid}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\keyword{variable}
