% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{speftotlt}
\alias{speftotlt}
\title{Total number of students enrolled in college spouse attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
speftotlt
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
J. Total Fall Enrollment
}
\section{Values}{

\itemize{
\item \code{1} 0-467 students enrolled
\item \code{2} 468 or more students
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/78/vshow}{GSS Data Explorer page for \code{speftotlt}}.
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
