% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{speth1}
\alias{speth1}
\title{1st mentioned country of spouse's origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
speth1
}
\details{
Question From what countries or part of the world did your (wife''s/husband''s) ancestors come?
B. FIRST ETHNICITY OF SPOUSE MENTIONED
}
\section{Values}{

\itemize{
\item \code{1} africa
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} rumania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span windies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5269/vshow}{GSS Data Explorer page for \code{speth1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa \tab austria \tab french canada \tab other canada \tab china \tab czechoslovakia \tab denmark \tab england & wales \tab finland \tab france \tab germany \tab greece \tab hungary \tab ireland \tab italy \tab japan \tab mexico \tab netherlands \tab norway \tab philippines \tab poland \tab puerto rico \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab other \tab american indian \tab india \tab portugal \tab lithuania \tab yugoslavia \tab rumania \tab arabic \tab other spanish \tab non-span windies \tab american only \tab don't know \tab no answer \tab west indies \tab belgium \tab other asian \tab other european \tab not available in this year \tab iap \tab Total \cr
   1985 \tab 663 \tab 24 \tab 9 \tab 10 \tab 6 \tab 2 \tab 10 \tab 10 \tab 114 \tab 6 \tab 26 \tab 148 \tab 5 \tab 2 \tab 97 \tab 38 \tab 4 \tab 26 \tab 17 \tab 17 \tab 2 \tab 31 \tab 5 \tab 13 \tab 23 \tab 4 \tab 21 \tab 5 \tab 6 \tab 32 \tab 1 \tab 3 \tab 3 \tab 2 \tab 3 \tab 1 \tab 3 \tab 2 \tab 9 \tab 127 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 644 \tab 33 \tab 6 \tab 7 \tab 4 \tab - \tab 8 \tab 8 \tab 108 \tab 4 \tab 14 \tab 145 \tab 2 \tab 7 \tab 91 \tab 40 \tab 4 \tab 24 \tab 14 \tab 15 \tab 1 \tab 22 \tab 8 \tab 14 \tab 31 \tab 6 \tab 16 \tab 3 \tab 5 \tab 21 \tab 5 \tab 4 \tab 3 \tab 6 \tab 1 \tab 3 \tab 3 \tab 4 \tab 13 \tab 120 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 903 \tab 84 \tab 8 \tab 12 \tab 2 \tab 5 \tab 14 \tab 12 \tab 92 \tab 1 \tab 26 \tab 127 \tab 1 \tab 4 \tab 68 \tab 43 \tab 2 \tab 20 \tab 20 \tab 14 \tab 1 \tab 37 \tab 1 \tab 15 \tab 32 \tab 6 \tab 14 \tab 3 \tab 11 \tab 35 \tab 2 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab 14 \tab 161 \tab 10 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 693 \tab 31 \tab 3 \tab 7 \tab 1 \tab 7 \tab 5 \tab 6 \tab 96 \tab 5 \tab 15 \tab 151 \tab 7 \tab 5 \tab 70 \tab 35 \tab 2 \tab 21 \tab 22 \tab 12 \tab - \tab 19 \tab 5 \tab 19 \tab 23 \tab 9 \tab 8 \tab 4 \tab 7 \tab 21 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 13 \tab 136 \tab 11 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 690 \tab 32 \tab 3 \tab 10 \tab 4 \tab 7 \tab 9 \tab 15 \tab 94 \tab 1 \tab 18 \tab 177 \tab 5 \tab 5 \tab 65 \tab 45 \tab 4 \tab 24 \tab 14 \tab 15 \tab 3 \tab 30 \tab 5 \tab 12 \tab 28 \tab 6 \tab 15 \tab 2 \tab 11 \tab 24 \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 6 \tab 125 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 645 \tab 33 \tab 7 \tab 8 \tab 3 \tab 1 \tab 10 \tab 6 \tab 96 \tab 3 \tab 17 \tab 117 \tab 2 \tab 2 \tab 69 \tab 36 \tab 3 \tab 23 \tab 16 \tab 19 \tab - \tab 23 \tab 7 \tab 10 \tab 24 \tab 9 \tab 8 \tab 7 \tab 6 \tab 23 \tab 1 \tab 5 \tab - \tab 3 \tab 2 \tab 1 \tab 3 \tab 6 \tab 6 \tab 104 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 713 \tab 45 \tab 3 \tab 13 \tab 11 \tab 3 \tab 8 \tab 13 \tab 110 \tab 2 \tab 19 \tab 128 \tab 2 \tab 5 \tab 85 \tab 45 \tab 4 \tab 25 \tab 13 \tab 8 \tab 3 \tab 23 \tab 4 \tab 12 \tab 26 \tab 6 \tab 12 \tab 2 \tab 10 \tab 24 \tab 3 \tab 2 \tab 5 \tab 2 \tab - \tab - \tab 3 \tab 6 \tab 6 \tab 105 \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 751 \tab 37 \tab 1 \tab 8 \tab 5 \tab 5 \tab 8 \tab 6 \tab 110 \tab 3 \tab 26 \tab 174 \tab 10 \tab 4 \tab 64 \tab 32 \tab 2 \tab 17 \tab 14 \tab 10 \tab 6 \tab 25 \tab 1 \tab 21 \tab 30 \tab 14 \tab 15 \tab 2 \tab 10 \tab 25 \tab 3 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 10 \tab 3 \tab 8 \tab 132 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2591 \tab 28 \tab 2 \tab 8 \tab 1 \tab - \tab 5 \tab 1 \tab 49 \tab 2 \tab 8 \tab 77 \tab 2 \tab 2 \tab 33 \tab 14 \tab 3 \tab 18 \tab 5 \tab 9 \tab 2 \tab 8 \tab 3 \tab 6 \tab 15 \tab - \tab 7 \tab 2 \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab 5 \tab 57 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 2992 \cr
   Total \tab 8293 \tab 347 \tab 42 \tab 83 \tab 37 \tab 30 \tab 77 \tab 77 \tab 869 \tab 27 \tab 169 \tab 1244 \tab 36 \tab 36 \tab 642 \tab 328 \tab 28 \tab 198 \tab 135 \tab 119 \tab 18 \tab 218 \tab 39 \tab 122 \tab 232 \tab 60 \tab 116 \tab 30 \tab 68 \tab 210 \tab 23 \tab 25 \tab 24 \tab 21 \tab 14 \tab 10 \tab 32 \tab 26 \tab 80 \tab 1067 \tab 63 \tab 1 \tab 7 \tab 4 \tab 1 \tab 0 \tab 0 \tab 15328 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}}
}
\concept{Ethnicity}
\keyword{variable}
