% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{speth2}
\alias{speth2}
\title{2nd mentioned country of spouse's origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
speth2
}
\details{
Question From what countries or part of the world did your (wife''s/husband''s) ancestors come?
C. SECOND ETHNICITY OF SPOUSE MENTIONED
}
\section{Values}{

\itemize{
\item \code{1} africa
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} rumania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span windies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5270/vshow}{GSS Data Explorer page for \code{speth2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa \tab austria \tab french canada \tab other canada \tab czechoslovakia \tab denmark \tab england & wales \tab finland \tab france \tab germany \tab hungary \tab ireland \tab italy \tab mexico \tab netherlands \tab norway \tab philippines \tab poland \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab other \tab american indian \tab portugal \tab lithuania \tab yugoslavia \tab arabic \tab no answer \tab puerto rico \tab belgium \tab american only \tab greece \tab india \tab rumania \tab non-span windies \tab other spanish \tab don't know \tab japan \tab not available in this year \tab iap \tab Total \cr
   1985 \tab 1247 \tab 2 \tab 1 \tab 4 \tab 3 \tab 4 \tab 2 \tab 50 \tab 2 \tab 16 \tab 44 \tab 3 \tab 55 \tab 6 \tab 2 \tab 5 \tab 15 \tab 1 \tab 5 \tab 5 \tab 25 \tab 4 \tab 6 \tab 1 \tab 4 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1191 \tab 1 \tab 3 \tab 5 \tab 5 \tab 5 \tab 2 \tab 48 \tab 2 \tab 13 \tab 42 \tab 3 \tab 54 \tab 10 \tab 2 \tab 7 \tab 6 \tab - \tab 7 \tab 1 \tab 24 \tab 3 \tab 4 \tab 6 \tab 8 \tab 9 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1543 \tab 1 \tab - \tab 2 \tab - \tab 5 \tab 1 \tab 39 \tab 2 \tab 23 \tab 48 \tab 7 \tab 33 \tab 8 \tab 5 \tab 11 \tab 5 \tab - \tab 8 \tab 8 \tab 23 \tab - \tab 7 \tab 3 \tab 4 \tab 12 \tab - \tab 2 \tab 2 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1245 \tab 1 \tab 3 \tab 4 \tab 3 \tab 2 \tab 1 \tab 37 \tab - \tab 13 \tab 32 \tab - \tab 47 \tab 4 \tab 2 \tab 9 \tab 4 \tab 1 \tab 10 \tab 1 \tab 13 \tab - \tab 11 \tab 3 \tab 4 \tab 15 \tab 1 \tab 1 \tab - \tab - \tab 12 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1263 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 49 \tab 2 \tab 13 \tab 38 \tab 1 \tab 65 \tab 3 \tab - \tab 6 \tab 7 \tab - \tab 9 \tab 2 \tab 21 \tab 3 \tab 3 \tab - \tab 3 \tab 15 \tab - \tab 1 \tab 2 \tab - \tab 15 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1130 \tab - \tab 1 \tab 3 \tab 3 \tab 5 \tab 1 \tab 26 \tab 1 \tab 15 \tab 45 \tab 2 \tab 43 \tab 6 \tab 3 \tab 7 \tab 5 \tab - \tab 8 \tab 4 \tab 15 \tab 2 \tab 11 \tab 2 \tab - \tab 18 \tab 1 \tab 2 \tab 3 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1297 \tab 1 \tab - \tab 7 \tab 2 \tab 1 \tab 6 \tab 27 \tab - \tab 8 \tab 33 \tab 1 \tab 43 \tab 11 \tab 2 \tab 9 \tab 4 \tab - \tab 14 \tab 4 \tab 12 \tab 2 \tab 12 \tab 3 \tab 1 \tab 12 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab 1322 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 2 \tab 36 \tab 1 \tab 15 \tab 43 \tab - \tab 67 \tab 10 \tab - \tab 11 \tab 2 \tab 1 \tab 13 \tab 3 \tab 16 \tab 10 \tab 8 \tab 5 \tab 3 \tab 14 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2825 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 17 \tab - \tab 6 \tab 20 \tab 1 \tab 22 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab 8 \tab 4 \tab 5 \tab 1 \tab 2 \tab 7 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 51 \tab 1 \tab - \tab - \tab 2992 \cr
   Total \tab 13063 \tab 9 \tab 11 \tab 33 \tab 20 \tab 29 \tab 19 \tab 329 \tab 10 \tab 122 \tab 345 \tab 18 \tab 429 \tab 61 \tab 17 \tab 68 \tab 49 \tab 3 \tab 76 \tab 30 \tab 157 \tab 28 \tab 67 \tab 24 \tab 29 \tab 116 \tab 4 \tab 10 \tab 13 \tab 2 \tab 54 \tab 4 \tab 2 \tab 3 \tab 4 \tab 2 \tab 8 \tab 2 \tab 3 \tab 53 \tab 2 \tab 0 \tab 0 \tab 15328 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{speth1}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}}
}
\concept{Ethnicity}
\keyword{variable}
