% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{speth3}
\alias{speth3}
\title{3rd mentioned country of spouse's origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
speth3
}
\details{
Question From what countries or part of the world did your (wife''s/husband''s) ancestors come?
D. THIRD ETHNICITY OF SPOUSE MENTIONED
}
\section{Values}{

\itemize{
\item \code{1} africa
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} rumania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span windies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5271/vshow}{GSS Data Explorer page for \code{speth3}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab denmark \tab england & wales \tab france \tab germany \tab ireland \tab italy \tab netherlands \tab norway \tab russia \tab scotland \tab sweden \tab switzerland \tab other \tab american indian \tab lithuania \tab belgium \tab no answer \tab other canada \tab czechoslovakia \tab india \tab yugoslavia \tab american only \tab austria \tab french canada \tab hungary \tab poland \tab spain \tab mexico \tab africa \tab japan \tab puerto rico \tab rumania \tab philippines \tab don't know \tab not available in this year \tab iap \tab Total \cr
   1985 \tab 1465 \tab 1 \tab 13 \tab 2 \tab 8 \tab 11 \tab 2 \tab 4 \tab 4 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1408 \tab - \tab 16 \tab 1 \tab 15 \tab 8 \tab 3 \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1746 \tab 1 \tab 7 \tab 6 \tab 5 \tab 6 \tab 3 \tab 5 \tab 1 \tab - \tab 4 \tab 5 \tab 4 \tab - \tab 4 \tab - \tab - \tab 10 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1411 \tab - \tab 12 \tab 4 \tab 10 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 2 \tab 8 \tab - \tab - \tab 12 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1464 \tab - \tab 12 \tab 2 \tab 9 \tab 9 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab - \tab 15 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1311 \tab 2 \tab 9 \tab 4 \tab 8 \tab 8 \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab 6 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1465 \tab 1 \tab 6 \tab 5 \tab 10 \tab 7 \tab - \tab 1 \tab 1 \tab 1 \tab 9 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1547 \tab 1 \tab 5 \tab 6 \tab 14 \tab 9 \tab 3 \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 9 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab 2917 \tab - \tab 2 \tab - \tab 4 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 50 \tab - \tab - \tab 2992 \cr
   Total \tab 14734 \tab 6 \tab 82 \tab 30 \tab 83 \tab 68 \tab 14 \tab 25 \tab 8 \tab 3 \tab 42 \tab 12 \tab 7 \tab 5 \tab 47 \tab 1 \tab 1 \tab 56 \tab 3 \tab 10 \tab 3 \tab 1 \tab 3 \tab 4 \tab 4 \tab 3 \tab 8 \tab 5 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 51 \tab 0 \tab 0 \tab 15328 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{spethnic}},
\code{\link{spethnum}}
}
\concept{Ethnicity}
\keyword{variable}
