% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{spfalook}
\alias{spfalook}
\title{Hours spouse spends looking after family members}
\source{
General Social Survey https://gss.norc.org
}
\description{
spfalook
}
\details{
Question And on average, how many hours a week does he/she spend looking after family members (e.g. children, elderly, ill or disabled family members)?
}
\section{Values}{

\itemize{
\item \code{1} 1 hour
\item \code{97} 97 hours or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4484/vshow}{GSS Data Explorer page for \code{spfalook}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{rfamlook}},
\code{\link{splive}},
\code{\link{wkyngscs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
