% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{spfam16}
\alias{spfam16}
\title{Did spouse live at home at age 16?}
\source{
General Social Survey https://gss.norc.org
}
\description{
spfam16
}
\details{
Question Was your (husband/wife) living with both (his/her) own mother and father around the time (s/he) was 16? (IF NO: with whom was (he/she) living around that time?) IF HUSBAND/WIFE MARRIED OR LEFT HOME BY AGE 16,
PROBE FOR BEFORE THAT.
}
\section{Values}{

\itemize{
\item \verb{[0]} other
\item \verb{[1]} both own mother and father
\item \verb{[2]} father and stepmother
\item \verb{[3]} mother and stepfather
\item \verb{[4]} father-no mother or stepmother
\item \verb{[5]} mother-no father or stepfather
\item \verb{[6]} some other male relative -no female head-
\item \verb{[7]} some other female relative -no male head-
\item \verb{[8]} other, with both male and female relatives
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1484/vshow}{GSS Data Explorer page for \code{spfam16}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab other \tab both own mother and father \tab father and stepmother \tab mother and stepfather \tab father-no mother or stepmother \tab mother-no father or stepfather \tab some other male relative -no female head- \tab some other female relative -no male head- \tab other, with both male and female relatives \tab no answer \tab Total \cr
   1994 \tab 2213 \tab 19 \tab 584 \tab 19 \tab 31 \tab 16 \tab 73 \tab 1 \tab 7 \tab 12 \tab 17 \tab 2992 \cr
   Total \tab 2213 \tab 19 \tab 584 \tab 19 \tab 31 \tab 16 \tab 73 \tab 1 \tab 7 \tab 12 \tab 17 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}
}
\concept{Family Mobility}
\keyword{variable}
