% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{spgradtounder}
\alias{spgradtounder}
\title{Ratio of graduate students to undergraduate students in college spouse attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
spgradtounder
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
L. Ratio of Graduate to Undergraduate students.
}
\section{Values}{

\itemize{
\item \code{1} 0-.210 graduate student per undergraduate student
\item \code{2} .211 or over
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/80/vshow}{GSS Data Explorer page for \code{spgradtounder}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 0-.210 graduate student per undergraduate student \tab .211 or over \tab no answer \tab Total \cr
   2012 \tab 1746 \tab 78 \tab 149 \tab 1 \tab 1974 \cr
   2014 \tab 2181 \tab 126 \tab 230 \tab 1 \tab 2538 \cr
   2016 \tab 2521 \tab 108 \tab 238 \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 8796 \tab 312 \tab 617 \tab 2 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}}
}
\concept{Core}
\keyword{variable}
