% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_158.R
\name{sphedoccindv}
\alias{sphedoccindv}
\title{Spouse's father's occ/ind verbatim retrieved}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedoccindv
}
\details{
Question Spouse's Head of Household OCC/IND verbatim recovered
}
\section{Values}{

\itemize{
\item \verb{[0]} no verbatims for sphed
\item \verb{[1]} verbatims for sphed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6002/vshow}{GSS Data Explorer page for \code{sphedoccindv}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no verbatims for sphed \tab verbatims for sphed \tab Total \cr
   1994 \tab 54 \tab 2174 \tab 764 \tab 2992 \cr
   Total \tab 54 \tab 2174 \tab 764 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc10}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre10}},
\code{\link{sphedslf}}
}
\concept{Family Mobility}
\keyword{variable}
