% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{sphedpre}
\alias{sphedpre}
\title{Spouse's father's occ pres when spouse was 16 (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedpre
}
\details{
Question A. Now we would like to find out what kind of work your (husband's/wife's) father (OR SUBSTITUTE) did when your (husband/wife) was about 16 years old. What kind of work was your (husband's/wife's) father (OR SUBSTITUTE) doing? That is, what was (his/her) job called?
B. IF NOT ALREADY ANSWERED, ASK:  What did (he/she) actually do in that job?  Tell me what were some of  (his/her) main duties?
C. What kind of place did (he/she) work for?
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was (he/she) self-employed or did (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1488/vshow}{GSS Data Explorer page for \code{sphedpre}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 53 \tab 54 \tab 57 \tab 58 \tab 60 \tab 61 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 71 \tab 72 \tab 74 \tab 75 \tab 86 \tab dk, na, iap \tab iap \tab Total \cr
   1994 \tab - \tab 1 \tab 11 \tab 14 \tab 29 \tab 5 \tab 2 \tab 5 \tab 15 \tab 8 \tab 22 \tab 7 \tab 13 \tab 7 \tab 15 \tab 22 \tab 29 \tab 1 \tab 4 \tab 31 \tab 101 \tab 7 \tab 20 \tab 7 \tab 29 \tab 12 \tab 4 \tab 21 \tab 8 \tab 26 \tab 7 \tab 59 \tab 5 \tab 4 \tab 1 \tab 2 \tab 8 \tab 6 \tab 17 \tab 6 \tab 13 \tab 3 \tab 6 \tab 4 \tab 5 \tab 2 \tab 2 \tab 2 \tab 2364 \tab - \tab 2992 \cr
   Total \tab 0 \tab 1 \tab 11 \tab 14 \tab 29 \tab 5 \tab 2 \tab 5 \tab 15 \tab 8 \tab 22 \tab 7 \tab 13 \tab 7 \tab 15 \tab 22 \tab 29 \tab 1 \tab 4 \tab 31 \tab 101 \tab 7 \tab 20 \tab 7 \tab 29 \tab 12 \tab 4 \tab 21 \tab 8 \tab 26 \tab 7 \tab 59 \tab 5 \tab 4 \tab 1 \tab 2 \tab 8 \tab 6 \tab 17 \tab 6 \tab 13 \tab 3 \tab 6 \tab 4 \tab 5 \tab 2 \tab 2 \tab 2 \tab 2364 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{spfam16}},
\code{\link{sphedocc}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}
}
\concept{Family Mobility}
\keyword{variable}
