% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_158.R
\name{sphedpre10}
\alias{sphedpre10}
\title{Spouse's father's occ pres when spouse was 16 (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedpre10
}
\details{
Question Spouse's father's prestige score when industry was 16. SPHEDPRE10 is based on the 2010 Census occupation classification. This standard prestige score is a simple mean value of ratings for each occupation category, converted to a scale of 0 (bottom) to 100 (top).
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6008/vshow}{GSS Data Explorer page for \code{sphedpre10}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 22 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 53 \tab 56 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 69 \tab 70 \tab 72 \tab 73 \tab 74 \tab 75 \tab 80 \tab Total \cr
   1994 \tab 2321 \tab 4 \tab 11 \tab 19 \tab 4 \tab 29 \tab 15 \tab 6 \tab 3 \tab 23 \tab 12 \tab 9 \tab 9 \tab 62 \tab 11 \tab 5 \tab 38 \tab 32 \tab 17 \tab 7 \tab 14 \tab 3 \tab 30 \tab 35 \tab 88 \tab 13 \tab 15 \tab 24 \tab 17 \tab 3 \tab 9 \tab 3 \tab 2 \tab 7 \tab 15 \tab 4 \tab 10 \tab 11 \tab 14 \tab 3 \tab 6 \tab 2 \tab 3 \tab 7 \tab 7 \tab 3 \tab 5 \tab 2 \tab 2992 \cr
   Total \tab 2321 \tab 4 \tab 11 \tab 19 \tab 4 \tab 29 \tab 15 \tab 6 \tab 3 \tab 23 \tab 12 \tab 9 \tab 9 \tab 62 \tab 11 \tab 5 \tab 38 \tab 32 \tab 17 \tab 7 \tab 14 \tab 3 \tab 30 \tab 35 \tab 88 \tab 13 \tab 15 \tab 24 \tab 17 \tab 3 \tab 9 \tab 3 \tab 2 \tab 7 \tab 15 \tab 4 \tab 10 \tab 11 \tab 14 \tab 3 \tab 6 \tab 2 \tab 3 \tab 7 \tab 7 \tab 3 \tab 5 \tab 2 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedslf}}
}
\concept{Family Mobility}
\keyword{variable}
