% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{sphmewrk}
\alias{sphmewrk}
\title{How much housework r's spouse does}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphmewrk
}
\details{
Question We're interested in knowing how much of the work around your home is done by different people. We don't mean taking care of children - just regular work around the house, like cooking, grocery shopping and doing little repair jobs.
B. How much does your spouse or partner do?
}
\section{Values}{

\itemize{
\item \code{1} all
\item \code{2} most
\item \code{3} about half
\item \code{4} some
\item \code{5} very little or none
\item \code{6} not married-living with partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1661/vshow}{GSS Data Explorer page for \code{sphmewrk}}.
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{rhmewrk}},
\code{\link{sharesep}},
\code{\link{tradmod}},
\code{\link{willwed1}},
\code{\link{willwed2}}

Other Household: 
\code{\link{rhmewrk}}
}
\concept{Gender}
\concept{Household}
\keyword{variable}
