% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{sphour}
\alias{sphour}
\title{Hours spouse usually works}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphour
}
\details{
Question What hours does (he/she) usually work--days, evenings, or nights?  CIRCLE ONE CODE.
}
\section{Values}{

\itemize{
\item \code{1} reg day
\item \code{2} eve
\item \code{3} nights
\item \code{4} varies
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5290/vshow}{GSS Data Explorer page for \code{sphour}}.

Counts by year:\tabular{lllllllll}{
   year \tab reg day \tab eve \tab nights \tab varies \tab other \tab iap \tab no answer \tab Total \cr
   1972 \tab 420 \tab 20 \tab 11 \tab 75 \tab 38 \tab 1044 \tab 5 \tab 1613 \cr
   Total \tab 420 \tab 20 \tab 11 \tab 75 \tab 38 \tab 1044 \tab 5 \tab 1613 \cr
}
}

\seealso{
Other Military Service & Work: 
\code{\link{spdays}},
\code{\link{vetkind}},
\code{\link{vetyears}},
\code{\link{workdy}},
\code{\link{workhr}}
}
\concept{Military Service & Work}
\keyword{variable}
