% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_04.R
\name{sphrs1}
\alias{sphrs1}
\title{Number of hrs spouse worked last week}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphrs1
}
\details{
Question Last week was your (wife/husband) working full time, part time, going to school, keeping house, or what?
A. IF WORKING, FULL OR PART TIME:  How many hours did (he/she) work last week, at all jobs?
}
\section{Values}{

\itemize{
\item \code{89} 89+ hours
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/24/vshow}{GSS Data Explorer page for \code{sphrs1}}.
}

\seealso{
Other Core: 
\code{\link{coevwork}},
\code{\link{cohrs1}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{divorce}},
\code{\link{spevwork}},
\code{\link{sphrs2}},
\code{\link{spwrksta}},
\code{\link{widowed}}

Other Hours Worked: 
\code{\link{sphrs2}}

Other Job: 
\code{\link{sphrs2}}

Other Labor Force Status: 
\code{\link{spevwork}},
\code{\link{sphrs2}},
\code{\link{spwrksta}}
}
\concept{Core}
\concept{Hours Worked}
\concept{Job}
\concept{Labor Force Status}
\keyword{variable}
