% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{spindtag}
\alias{spindtag}
\title{Spind10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
spindtag
}
\details{
Question SPIND10 Coding Tag
}
\section{Values}{

\itemize{
\item \code{1} found
\item \code{2} not found, assigned
\item \code{3} not found, not assigned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{coind10}},
\code{\link{coocc10}},
\code{\link{paocc16}},
\code{\link{spind10}},
\code{\link{spindstatus}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{sppres10}},
\code{\link{sppres105plus}}
}
\concept{Core}
\keyword{variable}
