% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{spjrel16}
\alias{spjrel16}
\title{Religion in which spouse raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
spjrel16
}
\details{
Question Religion in which spouse was raised.
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[6]} buddhism
\item \verb{[7]} hinduism
\item \verb{[8]} other eastern religions
\item \verb{[9]} muslim/islam
\item \verb{[10]} orthodox-christian
\item \verb{[11]} christian
\item \verb{[12]} native american
\item \verb{[13]} inter-nondenominational
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3824/vshow}{GSS Data Explorer page for \code{spjrel16}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab Total \cr
   2008 \tab 1997 \tab 8 \tab 3 \tab 15 \tab 2023 \cr
   Total \tab 1997 \tab 8 \tab 3 \tab 15 \tab 2023 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{majwoth}},
\code{\link{pajwoth}},
\code{\link{spjoth16}}
}
\concept{Jewish Identity}
\keyword{variable}
