% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{spkcom}
\alias{spkcom}
\title{Allow communist to speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
spkcom
}
\details{
Question Now, I should like to ask you some questions about a man who admits he is a Communist.
A. Suppose this admitted Communist wanted to make a speech in your community. Should he be allowed to speak, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes, allowed to speak
\item \code{2} not allowed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/255/vshow}{GSS Data Explorer page for \code{spkcom}}.

Counts by year:\tabular{llllllll}{
   year \tab yes, allowed to speak \tab not allowed \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 835 \tab 721 \tab 54 \tab 3 \tab - \tab - \tab 1613 \cr
   1973 \tab 898 \tab 570 \tab 35 \tab 1 \tab - \tab - \tab 1504 \cr
   1974 \tab 857 \tab 582 \tab 42 \tab 3 \tab - \tab - \tab 1484 \cr
   1976 \tab 817 \tab 647 \tab 33 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab 846 \tab 648 \tab 31 \tab 5 \tab - \tab - \tab 1530 \cr
   1980 \tab 809 \tab 621 \tab 37 \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 1002 \tab 769 \tab 81 \tab 8 \tab - \tab - \tab 1860 \cr
   1984 \tab 870 \tab 562 \tab 36 \tab 5 \tab - \tab - \tab 1473 \cr
   1985 \tab 872 \tab 616 \tab 41 \tab 5 \tab - \tab - \tab 1534 \cr
   1987 \tab 1066 \tab 706 \tab 43 \tab 4 \tab - \tab - \tab 1819 \cr
   1988 \tab 588 \tab 365 \tab 23 \tab 1 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 662 \tab 343 \tab 25 \tab 3 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 589 \tab 298 \tab 27 \tab 3 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 666 \tab 304 \tab 20 \tab 3 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 736 \tab 308 \tab 28 \tab 3 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 1331 \tab 619 \tab 42 \tab 4 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 1230 \tab 640 \tab 51 \tab 2 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 1249 \tab 577 \tab 50 \tab 6 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 1216 \tab 585 \tab 54 \tab 6 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 636 \tab 266 \tab 21 \tab 1 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 623 \tab 257 \tab 15 \tab 3 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 1339 \tab 623 \tab 39 \tab 2 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 874 \tab 442 \tab 34 \tab 2 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 819 \tab 423 \tab 33 \tab 6 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 862 \tab 395 \tab 44 \tab 7 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 1156 \tab 504 \tab 49 \tab 5 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 1300 \tab 523 \tab 52 \tab 15 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 1055 \tab 453 \tab 61 \tab 5 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 958 \tab 347 \tab 3 \tab - \tab 2714 \tab 10 \tab 4032 \cr
   2022 \tab 789 \tab 336 \tab 29 \tab 6 \tab 2380 \tab 4 \tab 3544 \cr
   Total \tab 27550 \tab 15050 \tab 1133 \tab 120 \tab 22432 \tab 14 \tab 66299 \cr
}
}

\seealso{
Other Core: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{libsoc}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Core}
\keyword{variable}
