% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{spkcomy}
\alias{spkcomy}
\title{Allow communist to speak y}
\source{
General Social Survey https://gss.norc.org
}
\description{
spkcomy
}
\details{
Question Now, we would like to ask you some questions about a self-identified Communist. Suppose this admitted Communist wanted to make a speech in your community. Should this person be allowed to speak, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, allowed to speak
\item \verb{[2]} not allowed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7438/vshow}{GSS Data Explorer page for \code{spkcomy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, allowed to speak \tab not allowed \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 2675 \tab 953 \tab 385 \tab 5 \tab 14 \tab - \tab 4032 \cr
   2022 \tab 2367 \tab 777 \tab 367 \tab 26 \tab 3 \tab 4 \tab 3544 \cr
   Total \tab 5042 \tab 1730 \tab 752 \tab 31 \tab 17 \tab 4 \tab 7576 \cr
}
}

\keyword{variable}
