% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{spklang}
\alias{spklang}
\title{How well does r speak other language}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklang
}
\details{
Question E. How well do you speak that language? (IF SPEAKS 2 OR MORE, ASK ONLY OF THE MOST FLUENT LANGUAGE) READ CATEGORIES
}
\section{Values}{

\itemize{
\item \verb{[1]} very well
\item \verb{[2]} well
\item \verb{[3]} not well
\item \verb{[4]} poorly/hardly at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2287/vshow}{GSS Data Explorer page for \code{spklang}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very well \tab well \tab not well \tab poorly/hardly at all \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2000 \tab 2450 \tab 137 \tab 85 \tab 92 \tab 40 \tab 13 \tab - \tab - \tab 2817 \cr
   2006 \tab 3837 \tab 272 \tab 127 \tab 197 \tab 75 \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab 1470 \tab 214 \tab 110 \tab 160 \tab 63 \tab 6 \tab - \tab - \tab 2023 \cr
   2010 \tab 1508 \tab 206 \tab 129 \tab 141 \tab 57 \tab 3 \tab - \tab - \tab 2044 \cr
   2012 \tab 1397 \tab 217 \tab 136 \tab 144 \tab 73 \tab 7 \tab - \tab - \tab 1974 \cr
   2014 \tab 1789 \tab 290 \tab 187 \tab 189 \tab 74 \tab 8 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 2045 \tab 296 \tab 207 \tab 203 \tab 105 \tab 11 \tab - \tab - \tab 2867 \cr
   2018 \tab 1625 \tab 260 \tab 181 \tab 192 \tab 86 \tab 4 \tab - \tab - \tab 2348 \cr
   2021 \tab 2810 \tab 441 \tab 275 \tab 305 \tab 45 \tab 146 \tab - \tab 10 \tab 4032 \cr
   2022 \tab 2554 \tab 439 \tab 231 \tab 205 \tab 52 \tab 61 \tab 1 \tab 1 \tab 3544 \cr
   2024 \tab 2404 \tab 420 \tab 206 \tab 202 \tab 60 \tab - \tab - \tab 17 \tab 3309 \cr
   Total \tab 23889 \tab 3192 \tab 1874 \tab 2030 \tab 730 \tab 261 \tab 2 \tab 28 \tab 32006 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{betrlang}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blkwrk}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{ethchng}},
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{knwwht}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{leastcom}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{uselang}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{uswht}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}

Other Language: 
\code{\link{betrlang}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{english}},
\code{\link{engoff1}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{uselang}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
