% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{spklang1}
\alias{spklang1}
\title{What language(s) do you speak well?}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklang1
}
\details{
Question What language(s) do you speak well?  PLEASE FILL IN.
First response
}
\section{Values}{

\itemize{
\item \code{1} english
\item \code{2} spanish
\item \code{3} polish
\item \code{4} french
\item \code{5} hebrew
\item \code{6} russian
\item \code{7} tagalog
\item \code{8} chinese
\item \code{9} urdu
\item \code{10} italian
\item \code{11} romanian
\item \code{12} german
\item \code{13} pashtu
\item \code{14} chamorro
\item \code{15} ukranian
\item \code{16} hungarian
\item \code{17} turkish
\item \code{18} khmer
\item \code{19} korean
\item \code{20} navajo
\item \code{21} malayam
\item \code{22} arabic
\item \code{23} farsi-persian
\item \code{24} serbian
\item \code{25} lithuanian
\item \code{27} thai
\item \code{28} gujarati
\item \code{30} dari
\item \code{31} slovak
\item \code{32} japanese
\item \code{33} portuguese
\item \code{34} vajun
\item \code{35} yidddish
\item \code{36} creole
\item \code{37} vietnamese
\item \code{38} norwegian
\item \code{39} malaysian
\item \code{40} latin
\item \code{41} czech
\item \code{42} hindu
\item \code{43} hakka
\item \code{44} swedish
\item \code{45} vesaya
\item \code{46} armenian
\item \code{47} american sign language
\item \code{48} cambodian
\item \code{49} spanglish
\item \code{50} tigrigna
\item \code{51} oriya
\item \code{52} hnong
\item \code{53} marathi
\item \code{54} wolof
\item \code{55} pangasian
\item \code{56} scottish
\item \code{57} macednian
\item \code{58} hawaiian
\item \code{59} fulani
\item \code{60} greek
\item \code{61} cajun
\item \code{62} akkadian
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4870/vshow}{GSS Data Explorer page for \code{spklang1}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{ethclose}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{parcit}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang2}},
\code{\link{spklang3}}

Other Language: 
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang2}},
\code{\link{spklang3}}
}
\concept{ISSP National Identity}
\concept{Language}
\keyword{variable}
