% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{spklang2}
\alias{spklang2}
\title{What language(s) do you speak well?}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklang2
}
\details{
Question What language(s) do you speak well?  PLEASE FILL IN.
Second response
}
\section{Values}{

\itemize{
\item \verb{[1]} english
\item \verb{[2]} spanish
\item \verb{[3]} polish
\item \verb{[4]} french
\item \verb{[5]} hebrew
\item \verb{[6]} russian
\item \verb{[7]} tagalog
\item \verb{[8]} chinese
\item \verb{[9]} urdu
\item \verb{[10]} italian
\item \verb{[11]} romanian
\item \verb{[12]} german
\item \verb{[13]} pashtu
\item \verb{[14]} chamorro
\item \verb{[15]} ukranian
\item \verb{[16]} hungarian
\item \verb{[17]} turkish
\item \verb{[18]} khmer
\item \verb{[19]} korean
\item \verb{[20]} navajo
\item \verb{[21]} malayam
\item \verb{[22]} arabic
\item \verb{[23]} farsi-persian
\item \verb{[24]} serbian
\item \verb{[25]} lithuanian
\item \verb{[27]} thai
\item \verb{[28]} gujarati
\item \verb{[30]} dari
\item \verb{[31]} slovak
\item \verb{[32]} japanese
\item \verb{[33]} portuguese
\item \verb{[34]} vajun
\item \verb{[35]} yidddish
\item \verb{[36]} creole
\item \verb{[37]} vietnamese
\item \verb{[38]} norwegian
\item \verb{[39]} malaysian
\item \verb{[40]} latin
\item \verb{[41]} czech
\item \verb{[42]} hindu
\item \verb{[43]} hakka
\item \verb{[44]} swedish
\item \verb{[45]} vesaya
\item \verb{[46]} armenian
\item \verb{[47]} american sign language
\item \verb{[48]} cambodian
\item \verb{[49]} spanglish
\item \verb{[50]} tigrigna
\item \verb{[51]} oriya
\item \verb{[52]} hnong
\item \verb{[53]} marathi
\item \verb{[54]} wolof
\item \verb{[55]} pangasian
\item \verb{[56]} scottish
\item \verb{[57]} macednian
\item \verb{[58]} hawaiian
\item \verb{[59]} fulani
\item \verb{[60]} greek
\item \verb{[61]} cajun
\item \verb{[62]} akkadian
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4871/vshow}{GSS Data Explorer page for \code{spklang2}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab english \tab spanish \tab polish \tab french \tab tagalog \tab urdu \tab italian \tab german \tab hungarian \tab korean \tab navajo \tab arabic \tab farsi-persian \tab lithuanian \tab gujarati \tab yidddish \tab creole \tab vietnamese \tab norwegian \tab malaysian \tab latin \tab czech \tab hindu \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 2757 \tab 38 \tab 45 \tab 3 \tab 16 \tab 3 \tab 2 \tab 6 \tab 16 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2904 \cr
   Total \tab 2757 \tab 38 \tab 45 \tab 3 \tab 16 \tab 3 \tab 2 \tab 6 \tab 16 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amproud1}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{belikeus}},
\code{\link{biasnews}},
\code{\link{bizmoney}},
\code{\link{brneffrt}},
\code{\link{citizen}},
\code{\link{citworld}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{decsorgs}},
\code{\link{ethclose}},
\code{\link{experts}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{freetrde}},
\code{\link{fulldem}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immassim}},
\code{\link{immcosts}},
\code{\link{immcrime1}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{immideas}},
\code{\link{immideas1}},
\code{\link{immimp}},
\code{\link{immjobs1}},
\code{\link{immrghts}},
\code{\link{immwlfare}},
\code{\link{imports}},
\code{\link{incgap1}},
\code{\link{internet}},
\code{\link{intlincs}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{lessprd}},
\code{\link{letin1}},
\code{\link{letin1a}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{refugees}},
\code{\link{res161}},
\code{\link{shortcom}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang3}},
\code{\link{wrldgovt}}

Other Language: 
\code{\link{betrlang}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{english}},
\code{\link{engoff1}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang1_0022}},
\code{\link{othlang2}},
\code{\link{othlang2_0022}},
\code{\link{parlang}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang}},
\code{\link{spklang1}},
\code{\link{spklang3}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{uselang}}
}
\concept{ISSP National Identity}
\concept{Language}
\keyword{variable}
