% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{spklang3}
\alias{spklang3}
\title{What language(s) do you speak well?}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklang3
}
\details{
Question What language(s) do you speak well?  PLEASE FILL IN.
Third response
}
\section{Values}{

\itemize{
\item \verb{[1]} english
\item \verb{[2]} spanish
\item \verb{[3]} polish
\item \verb{[4]} french
\item \verb{[5]} hebrew
\item \verb{[6]} russian
\item \verb{[7]} tagalog
\item \verb{[8]} chinese
\item \verb{[9]} urdu
\item \verb{[10]} italian
\item \verb{[11]} romanian
\item \verb{[12]} german
\item \verb{[13]} pashtu
\item \verb{[14]} chamorro
\item \verb{[15]} ukranian
\item \verb{[16]} hungarian
\item \verb{[17]} turkish
\item \verb{[18]} khmer
\item \verb{[19]} korean
\item \verb{[20]} navajo
\item \verb{[21]} malayam
\item \verb{[22]} arabic
\item \verb{[23]} farsi-persian
\item \verb{[24]} serbian
\item \verb{[25]} lithuanian
\item \verb{[27]} thai
\item \verb{[28]} gujarati
\item \verb{[30]} dari
\item \verb{[31]} slovak
\item \verb{[32]} japanese
\item \verb{[33]} portuguese
\item \verb{[34]} vajun
\item \verb{[35]} yidddish
\item \verb{[36]} creole
\item \verb{[37]} vietnamese
\item \verb{[38]} norwegian
\item \verb{[39]} malaysian
\item \verb{[40]} latin
\item \verb{[41]} czech
\item \verb{[42]} hindu
\item \verb{[43]} hakka
\item \verb{[44]} swedish
\item \verb{[45]} vesaya
\item \verb{[46]} armenian
\item \verb{[47]} american sign language
\item \verb{[48]} cambodian
\item \verb{[49]} spanglish
\item \verb{[50]} tigrigna
\item \verb{[51]} oriya
\item \verb{[52]} hnong
\item \verb{[53]} marathi
\item \verb{[54]} wolof
\item \verb{[55]} pangasian
\item \verb{[56]} scottish
\item \verb{[57]} macednian
\item \verb{[58]} hawaiian
\item \verb{[59]} fulani
\item \verb{[60]} greek
\item \verb{[61]} cajun
\item \verb{[62]} akkadian
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4872/vshow}{GSS Data Explorer page for \code{spklang3}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab english \tab spanish \tab french \tab russian \tab italian \tab german \tab japanese \tab portuguese \tab hindu \tab hakka \tab no answer \tab Total \cr
   1996 \tab 2879 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2904 \cr
   Total \tab 2879 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{ethclose}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{parcit}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}}
}
\concept{ISSP National Identity}
\keyword{variable}
