% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{spklangw}
\alias{spklangw}
\title{How well r speaks the language in betrlang}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklangw
}
\details{
Question C. How well do you speak that language?
}
\section{Values}{

\itemize{
\item \verb{[1]} very well
\item \verb{[2]} well
\item \verb{[3]} not well
\item \verb{[4]} poorly/hardly at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2303/vshow}{GSS Data Explorer page for \code{spklangw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very well \tab well \tab not well \tab poorly/hardly at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 4320 \tab 37 \tab 44 \tab 57 \tab 40 \tab 8 \tab 4 \tab 4510 \cr
   Total \tab 4320 \tab 37 \tab 44 \tab 57 \tab 40 \tab 8 \tab 4 \tab 4510 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{othhome}},
\code{\link{uswht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
