% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{spkmslm}
\alias{spkmslm}
\title{Allow muslim clergymen preaching hatred of the us}
\source{
General Social Survey https://gss.norc.org
}
\description{
spkmslm
}
\details{
Question Now consider a Muslim clergyman who preaches hatred of the United States.
A. If such a person wanted to make a speech in your community preaching hatred of the United States, should he be allowed  to speak, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, allowed to speak
\item \verb{[2]} not allowed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/264/vshow}{GSS Data Explorer page for \code{spkmslm}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, allowed to speak \tab not allowed \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 671 \tab 567 \tab 761 \tab 16 \tab 8 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 528 \tab 736 \tab 14 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 569 \tab 713 \tab 19 \tab 7 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 748 \tab 945 \tab 14 \tab 7 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 825 \tab 1031 \tab 20 \tab 14 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 715 \tab 832 \tab 18 \tab 9 \tab - \tab 2348 \cr
   2021 \tab 2714 \tab 629 \tab 678 \tab 1 \tab - \tab 10 \tab 4032 \cr
   2022 \tab 2380 \tab 540 \tab 601 \tab 12 \tab 5 \tab 6 \tab 3544 \cr
   Total \tab 9769 \tab 5121 \tab 6297 \tab 114 \tab 53 \tab 16 \tab 21370 \cr
}
}

\seealso{
Other Core: 
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}}
}
\concept{Core}
\keyword{variable}
