% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{spkracy}
\alias{spkracy}
\title{Allow anti-religious book in library y}
\source{
General Social Survey https://gss.norc.org
}
\description{
spkracy
}
\details{
Question Or consider a person who believes that Blacks are genetically inferior… If such a person wanted to make a speech in your community claiming that Blacks are inferior, should this person be allowed to speak, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes, allowed to speak
\item \code{2} not allowed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7436/vshow}{GSS Data Explorer page for \code{spkracy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, allowed to speak \tab not allowed \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2675 \tab 626 \tab 716 \tab 2 \tab 1 \tab 12 \tab 4032 \cr
   2022 \tab 2367 \tab 569 \tab 591 \tab 11 \tab 5 \tab 1 \tab 3544 \cr
   Total \tab 5042 \tab 1195 \tab 1307 \tab 13 \tab 6 \tab 13 \tab 7576 \cr
}
}

\keyword{variable}
