% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{spksoc}
\alias{spksoc}
\title{Allow socialist to speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
spksoc
}
\details{
Question Or consider a person who favored government ownership of all the railroads and all big industries.
A. If such a person wanted to make a speech in your community favoring government ownership of all the railroads and big industries, should he be allowed to speak, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, allowed to speak
\item \verb{[2]} not allowed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973, 1974 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/249/vshow}{GSS Data Explorer page for \code{spksoc}}.

Counts by year:\tabular{lllllll}{
   year \tab yes, allowed to speak \tab not allowed \tab don't know \tab no answer \tab iap \tab Total \cr
   1972 \tab 1241 \tab 293 \tab 78 \tab 1 \tab - \tab 1613 \cr
   1973 \tab 1165 \tab 296 \tab 43 \tab - \tab - \tab 1504 \cr
   1974 \tab 1122 \tab 307 \tab 54 \tab 1 \tab - \tab 1484 \cr
   Total \tab 3528 \tab 896 \tab 175 \tab 2 \tab 0 \tab 4601 \cr
}
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{tax}}
}
\concept{Core}
\keyword{variable}
