% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{spmarnum}
\alias{spmarnum}
\title{Has current spouse been married before?}
\source{
General Social Survey https://gss.norc.org
}
\description{
spmarnum
}
\details{
Question Has your (husband/wife) been married more than once?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1483/vshow}{GSS Data Explorer page for \code{spmarnum}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1994 \tab 2213 \tab 171 \tab 576 \tab 32 \tab 2992 \cr
   Total \tab 2213 \tab 171 \tab 576 \tab 32 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{spfam16}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}
}
\concept{Family Mobility}
\keyword{variable}
