% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{spmentl}
\alias{spmentl}
\title{Govt spending on health}
\source{
General Social Survey https://gss.norc.org
}
\description{
spmentl
}
\details{
Question Earlier, we talked about various areas of government spending. Since we've been talking about the mental health area, please indicate whether you would like to see more or less government spending in the area of mental health care. Remember that if you say "much more," it might require a tax increase to pay for it.
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend more
\item \verb{[3]} spend the same as now
\item \verb{[4]} spend less
\item \verb{[5]} spend much less
\item \verb{[8]} cant choose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2038/vshow}{GSS Data Explorer page for \code{spmentl}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend the same as now \tab spend less \tab spend much less \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 2168 \tab 86 \tab 248 \tab 267 \tab 61 \tab 15 \tab 48 \tab 11 \tab - \tab 2904 \cr
   2006 \tab 3775 \tab 79 \tab 276 \tab 299 \tab 36 \tab 14 \tab 24 \tab 7 \tab - \tab 4510 \cr
   Total \tab 5943 \tab 165 \tab 524 \tab 566 \tab 97 \tab 29 \tab 72 \tab 18 \tab 0 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{avoidmh}},
\code{\link{breakdwn}},
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{brkddo1}},
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{brkdtime}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{charactr}},
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_oth}},
\code{\link{chldshop_wht}},
\code{\link{chldvig}},
\code{\link{cutself}},
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{depress}},
\code{\link{diagnosd}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{doclist}},
\code{\link{docsui}},
\code{\link{dofirst}},
\code{\link{evbrkdwn}},
\code{\link{evmhp}},
\code{\link{fammhneg}},
\code{\link{fekids5}},
\code{\link{fighpeer}},
\code{\link{foodalc}},
\code{\link{genetics}},
\code{\link{gesttalk}},
\code{\link{godswill}},
\code{\link{govmentl}},
\code{\link{hlthdays}},
\code{\link{hlthinsr}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hospice}},
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imbalnce}},
\code{\link{imprvown}},
\code{\link{imprvown1}},
\code{\link{imprvtrt}},
\code{\link{instrctn}},
\code{\link{knwmhosp}},
\code{\link{knwnproz}},
\code{\link{knwpatnt}},
\code{\link{medaddct}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{meddoc}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentlill}},
\code{\link{mentloth}},
\code{\link{mhaccptd}},
\code{\link{mhbrain}},
\code{\link{mhcause}},
\code{\link{mhchrctr}},
\code{\link{mhclsoth}},
\code{\link{mhcreatv}},
\code{\link{mhdecslf}},
\code{\link{mhdoactv}},
\code{\link{mhdodiet}},
\code{\link{mhdodoc}},
\code{\link{mhdoexrc}},
\code{\link{mhdofam}},
\code{\link{mhdofrgt}},
\code{\link{mhdofrnd}},
\code{\link{mhdogrp}},
\code{\link{mhdoheal}},
\code{\link{mhdoherb}},
\code{\link{mhdohosp}},
\code{\link{mhdolife}},
\code{\link{mhdomhp}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdootc}},
\code{\link{mhdooth}},
\code{\link{mhdopray}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhdorx}},
\code{\link{mhexpoth}},
\code{\link{mhfamshm}},
\code{\link{mhfeelem}},
\code{\link{mhfriend}},
\code{\link{mhfright}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhgvtjob}},
\code{\link{mhhired}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlpoth}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhhrdtlk}},
\code{\link{mhhsptrt}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhkdcare}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhluck}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhmoney}},
\code{\link{mhneihbr}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhnotch}},
\code{\link{mhnotell}},
\code{\link{mhnotpol}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhoutsdr}},
\code{\link{mhphyscl}},
\code{\link{mhprdctv}},
\code{\link{mhproblm}},
\code{\link{mhraised}},
\code{\link{mhreloth}},
\code{\link{mhresoth}},
\code{\link{mhsecret}},
\code{\link{mhseedoc}},
\code{\link{mhseepub}},
\code{\link{mhseroth}},
\code{\link{mhslfshm}},
\code{\link{mhsocial}},
\code{\link{mhstress}},
\code{\link{mhsupwrk}},
\code{\link{mhsymp}},
\code{\link{mhtrtot2}},
\code{\link{mhtrtoth}},
\code{\link{mhtrtslf}},
\code{\link{mhtrusty}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}},
\code{\link{mhupdown}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}},
\code{\link{misswork}},
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}},
\code{\link{mnilwhy3}},
\code{\link{mntlas1}},
\code{\link{mntlas2}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{mntlhlth}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlrel}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}},
\code{\link{morextrm}},
\code{\link{mustdoc}},
\code{\link{mustdoc1}},
\code{\link{musthosp}},
\code{\link{musthosp1}},
\code{\link{mustmed}},
\code{\link{mustmed1}},
\code{\link{negself}},
\code{\link{notreat}},
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}},
\code{\link{otcmed}},
\code{\link{othmhneg}},
\code{\link{painmed}},
\code{\link{physill}},
\code{\link{physill1}},
\code{\link{probfix1}},
\code{\link{probfix2}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozac}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozkid3}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}},
\code{\link{pryntfam}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psyckid3}},
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{psycpsyc}},
\code{\link{putsoff}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}},
\code{\link{rightref}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{rxmed}},
\code{\link{schshoot}},
\code{\link{seemhpub}},
\code{\link{seenmntl}},
\code{\link{selfhelp}},
\code{\link{seriousp}},
\code{\link{smedia}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sphealer}},
\code{\link{stresses}},
\code{\link{swithlth}},
\code{\link{takelife}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}},
\code{\link{termvig}},
\code{\link{tlkclrgy}},
\code{\link{tlkfam}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}},
\code{\link{upsdowns}},
\code{\link{usedmntl}},
\code{\link{usedproz}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{viglab06}},
\code{\link{viglabel}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigvermy}},
\code{\link{vigversn}},
\code{\link{vigwork}},
\code{\link{wayraise}}

Other Federal Government: 
\code{\link{advfront}},
\code{\link{affrmact}},
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{concong}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{congeth}},
\code{\link{congovt}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{demworks}},
\code{\link{equalize}},
\code{\link{fedtrust}},
\code{\link{govcheat}},
\code{\link{govtpow}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{hlthcare}},
\code{\link{hspgovt}},
\code{\link{immcosts}},
\code{\link{immfare}},
\code{\link{jobaff}},
\code{\link{jobsall}},
\code{\link{knomentl}},
\code{\link{leadsci}},
\code{\link{mustret}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdiscm}},
\code{\link{natdisin}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natrecon}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{owneth}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{poleff17}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{pricecon}},
\code{\link{reqinfo}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{scresrch}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secterr}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{spnatdis}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{taxcheat}},
\code{\link{taxspend}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{whtgovt}}

Other Mental Health: 
\code{\link{anydoc}},
\code{\link{avoidmh}},
\code{\link{breakdwn}},
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{brkddo1}},
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{brkdtime}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{charactr}},
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_oth}},
\code{\link{chldshop_wht}},
\code{\link{chldvig}},
\code{\link{cutself}},
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{depress}},
\code{\link{diagnosd}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{doclist}},
\code{\link{docsui}},
\code{\link{dofirst}},
\code{\link{evbrkdwn}},
\code{\link{evmhp}},
\code{\link{fammhneg}},
\code{\link{fekids5}},
\code{\link{fighpeer}},
\code{\link{foodalc}},
\code{\link{genetics}},
\code{\link{gesttalk}},
\code{\link{godswill}},
\code{\link{govmentl}},
\code{\link{hlthdays}},
\code{\link{hlthinsr}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hospice}},
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imbalnce}},
\code{\link{imprvown}},
\code{\link{imprvown1}},
\code{\link{imprvtrt}},
\code{\link{instrctn}},
\code{\link{knwmhosp}},
\code{\link{knwnproz}},
\code{\link{knwpatnt}},
\code{\link{medaddct}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{meddoc}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentlill}},
\code{\link{mentloth}},
\code{\link{mhaccptd}},
\code{\link{mhbrain}},
\code{\link{mhcause}},
\code{\link{mhchrctr}},
\code{\link{mhclsoth}},
\code{\link{mhcreatv}},
\code{\link{mhdecslf}},
\code{\link{mhdoactv}},
\code{\link{mhdodiet}},
\code{\link{mhdodoc}},
\code{\link{mhdoexrc}},
\code{\link{mhdofam}},
\code{\link{mhdofrgt}},
\code{\link{mhdofrnd}},
\code{\link{mhdogrp}},
\code{\link{mhdoheal}},
\code{\link{mhdoherb}},
\code{\link{mhdohosp}},
\code{\link{mhdolife}},
\code{\link{mhdomhp}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdootc}},
\code{\link{mhdooth}},
\code{\link{mhdopray}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhdorx}},
\code{\link{mhexpoth}},
\code{\link{mhfamshm}},
\code{\link{mhfeelem}},
\code{\link{mhfriend}},
\code{\link{mhfright}},
\code{\link{mhgenes}},
\code{\link{mhgod}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthlt}},
\code{\link{mhgvthme}},
\code{\link{mhgvtjob}},
\code{\link{mhhired}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlpoth}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhhrdtlk}},
\code{\link{mhhsptrt}},
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhkdcare}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhluck}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhmoney}},
\code{\link{mhneihbr}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhnotch}},
\code{\link{mhnotell}},
\code{\link{mhnotpol}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhoutsdr}},
\code{\link{mhphyscl}},
\code{\link{mhprdctv}},
\code{\link{mhproblm}},
\code{\link{mhraised}},
\code{\link{mhreloth}},
\code{\link{mhresoth}},
\code{\link{mhsecret}},
\code{\link{mhseedoc}},
\code{\link{mhseepub}},
\code{\link{mhseroth}},
\code{\link{mhslfshm}},
\code{\link{mhsocial}},
\code{\link{mhstress}},
\code{\link{mhsupwrk}},
\code{\link{mhsymp}},
\code{\link{mhtrtot2}},
\code{\link{mhtrtoth}},
\code{\link{mhtrtslf}},
\code{\link{mhtrusty}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}},
\code{\link{mhupdown}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}},
\code{\link{misswork}},
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}},
\code{\link{mnilwhy3}},
\code{\link{mntlas1}},
\code{\link{mntlas2}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{mntlhlth}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlrel}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}},
\code{\link{morextrm}},
\code{\link{mustdoc}},
\code{\link{mustdoc1}},
\code{\link{musthosp}},
\code{\link{musthosp1}},
\code{\link{mustmed}},
\code{\link{mustmed1}},
\code{\link{negself}},
\code{\link{notreat}},
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}},
\code{\link{otcmed}},
\code{\link{othmhneg}},
\code{\link{painmed}},
\code{\link{physill}},
\code{\link{physill1}},
\code{\link{probfix1}},
\code{\link{probfix2}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozac}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozkid3}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}},
\code{\link{pryntfam}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psyckid3}},
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{psycpsyc}},
\code{\link{putsoff}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}},
\code{\link{rightref}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{rxmed}},
\code{\link{schshoot}},
\code{\link{seemhpub}},
\code{\link{seenmntl}},
\code{\link{selfhelp}},
\code{\link{seriousp}},
\code{\link{smedia}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sphealer}},
\code{\link{stresses}},
\code{\link{swithlth}},
\code{\link{takelife}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}},
\code{\link{termvig}},
\code{\link{tlkclrgy}},
\code{\link{tlkfam}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}},
\code{\link{upsdowns}},
\code{\link{usedmntl}},
\code{\link{usedproz}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{viglab06}},
\code{\link{viglabel}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigvermy}},
\code{\link{vigversn}},
\code{\link{vigwork}},
\code{\link{wayraise}}

Other Spending: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{cutdebt}},
\code{\link{cutgovt}},
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{natdrugz}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{spenviro}},
\code{\link{sphlth}},
\code{\link{spnatdis}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}},
\code{\link{taxspend}}
}
\concept{Federal Government}
\concept{Mental Health}
\concept{Spending}
\keyword{variable}
