% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{sppaid}
\alias{sppaid}
\title{Regularly received alimony - chld supprt}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppaid
}
\details{
Question Are you currently or have you ever been entitled to alimony or child support payments from a former (husband/wife)?
A. During the time you were entitled to alimony or child support, did you receive these payments on a regular basis, occasionally, or not at all?
}
\section{Values}{

\itemize{
\item \code{1} regular basis
\item \code{2} occasionally
\item \code{3} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/630/vshow}{GSS Data Explorer page for \code{sppaid}}.
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{spdue}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\keyword{variable}
