% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{sppart}
\alias{sppart}
\title{R's current insurance plan covers spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppart
}
\details{
Question Are the following people in your family covered by your current insurance plan?
A. My (spouse/partner)
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} not applicable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3932/vshow}{GSS Data Explorer page for \code{sppart}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab not applicable \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 1888 \tab 26 \tab 27 \tab 39 \tab 43 \tab - \tab 2023 \cr
   Total \tab 1888 \tab 26 \tab 27 \tab 39 \tab 43 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{dwell5}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evkid}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{mykids}},
\code{\link{negjob5}},
\code{\link{othkids}},
\code{\link{relgendr}},
\code{\link{sexsex18}},
\code{\link{sexunion}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}}

Other Family: 
\code{\link{adoptkid}},
\code{\link{agewed}},
\code{\link{agewedcr}},
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{balwkfam}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{chldben}},
\code{\link{chlddth}},
\code{\link{chldhrt}},
\code{\link{clsclsex}},
\code{\link{cousins}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{divorce}},
\code{\link{engda}},
\code{\link{engson}},
\code{\link{evdiv}},
\code{\link{evkid}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{famlife}},
\code{\link{famper1}},
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famvswk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{feovrwrk}},
\code{\link{flextime}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{inlawnum}},
\code{\link{inperson}},
\code{\link{jobvsfam}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kidcall}},
\code{\link{kidlived}},
\code{\link{kidmost}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{kidvisit}},
\code{\link{letters}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{madeath}},
\code{\link{malive}},
\code{\link{malive1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{marital}},
\code{\link{marunion}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{meetings}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{mykids}},
\code{\link{niecenep}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{numcntct}},
\code{\link{numprobs}},
\code{\link{obtohelp}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{othkids}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{padeath}},
\code{\link{palive}},
\code{\link{palive1}},
\code{\link{parleave}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{pawrkgrw}},
\code{\link{payrborn}},
\code{\link{payrdied}},
\code{\link{posslq}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relcall}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relmarry}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{rfamlook}},
\code{\link{sibcall}},
\code{\link{sibdeath}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibvisit}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{spdeath}},
\code{\link{spfalook}},
\code{\link{spfam16}},
\code{\link{spjoth16}},
\code{\link{spjrel16}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}},
\code{\link{widowed}},
\code{\link{wkvsfam}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}}

Other Health: 
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{anscitst}},
\code{\link{anydoc}},
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{crelesse}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{diffcare}},
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld4}},
\code{\link{disabld5}},
\code{\link{disabld6}},
\code{\link{disabld7}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc15a}},
\code{\link{doc16}},
\code{\link{doc16a}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc18a}},
\code{\link{doc19}},
\code{\link{doc19a}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc20a}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{docaskme}},
\code{\link{doccosts}},
\code{\link{docdecid}},
\code{\link{docinfo}},
\code{\link{doclist}},
\code{\link{docnomri}},
\code{\link{docrely}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{docvig}},
\code{\link{downblue}},
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downfeel}},
\code{\link{downqol}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotfeel}},
\code{\link{emotqol}},
\code{\link{emotrely}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{energy}},
\code{\link{findnwjb}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{health}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{helpsick}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hlthall}},
\code{\link{hlthcare}},
\code{\link{hlthcovr}},
\code{\link{hlthdays}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthinsr}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthplan}},
\code{\link{hlthtv}},
\code{\link{hlthtype}},
\code{\link{hlthwww}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hrdshp6}},
\code{\link{hyperten}},
\code{\link{insrchng}},
\code{\link{insrlmts}},
\code{\link{intmed}},
\code{\link{limitedp}},
\code{\link{misswork}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{modact}},
\code{\link{mykids}},
\code{\link{nodocidu}},
\code{\link{othkids}},
\code{\link{othplan}},
\code{\link{pain}},
\code{\link{painarms}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}},
\code{\link{peaceful}},
\code{\link{physhlth}},
\code{\link{regdoc}},
\code{\link{regdocyr}},
\code{\link{sathealt}},
\code{\link{secopin}},
\code{\link{sexpaid}},
\code{\link{slpprblm}},
\code{\link{socacts}},
\code{\link{stairs}},
\code{\link{swithlth}},
\code{\link{treat1}},
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{usedmntl}},
\code{\link{usedup}},
\code{\link{wait1}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}},
\code{\link{weight}}
}
\concept{Family}
\concept{Health}
\concept{Sexual Orientation}
\keyword{variable}
