% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_560.R
\name{sppill}
\alias{sppill}
\title{Spending for birth control}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppill
}
\details{
Question Here are some other areas of government spending. Please indicate whether you would like to see more or less government spending in each area.
Remember that if you say "much more", it might require a tax increase to pay for it.
J. Contraceptive services for teenagers…
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend more
\item \verb{[3]} spend same
\item \verb{[4]} spend less
\item \verb{[5]} spend much less
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5131/vshow}{GSS Data Explorer page for \code{sppill}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend same \tab spend less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1990 \tab 155 \tab 179 \tab 278 \tab 377 \tab 123 \tab 157 \tab 75 \tab 28 \tab 1372 \cr
   Total \tab 155 \tab 179 \tab 278 \tab 377 \tab 123 \tab 157 \tab 75 \tab 28 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppoorkd}},
\code{\link{spwrkpar}}
}
\concept{Children}
\keyword{variable}
