% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{sppres10}
\alias{sppres10}
\title{Spouse occupational prestige score (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppres10
}
\details{
Question SPPRES10 is based on the 2010 Census occupation classification. This standard prestige score is a simple mean value of ratings for each occupation category, converted to a scale of 0 (bottom) to 100 (top).
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{coind10}},
\code{\link{coocc10}},
\code{\link{paocc16}},
\code{\link{spind10}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{sppres105plus}}
}
\concept{Core}
\keyword{variable}
