% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{sppres80}
\alias{sppres80}
\title{Spouse's occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppres80
}
\details{
Question Spouse's Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/-/- \tab 2021 \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
}
}

\seealso{
Other Core: 
\code{\link{cohrs2}},
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{sppres}},
\code{\link{spwrkslf}}

Other Occupation: 
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{spwrkslf}}

Other Occupational Prestige: 
\code{\link{sppres}}
}
\concept{Core}
\concept{Occupation}
\concept{Occupational Prestige}
\keyword{variable}
