% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_633.R
\name{sprtconnct}
\alias{sprtconnct}
\title{How often r felt connected to world around them}
\source{
General Social Survey https://gss.norc.org
}
\description{
sprtconnct
}
\details{
Question How often R felt connected to world around them
}
\section{Values}{

\itemize{
\item \verb{[1]} at least once a day
\item \verb{[2]} almost every day
\item \verb{[3]} once or twice a week
\item \verb{[4]} once or twice a month
\item \verb{[5]} a few times per year
\item \verb{[6]} once a year or less
\item \verb{[7]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7534/vshow}{GSS Data Explorer page for \code{sprtconnct}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab at least once a day \tab almost every day \tab once or twice a week \tab once or twice a month \tab a few times per year \tab once a year or less \tab never \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 411 \tab 324 \tab 828 \tab 514 \tab 503 \tab 654 \tab 279 \tab 421 \tab 8 \tab 7 \tab 83 \tab 4032 \cr
   Total \tab 411 \tab 324 \tab 828 \tab 514 \tab 503 \tab 654 \tab 279 \tab 421 \tab 8 \tab 7 \tab 83 \tab 4032 \cr
}
}

\keyword{variable}
