% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_442.R
\name{spschool}
\alias{spschool}
\title{Govt spending on education}
\source{
General Social Survey https://gss.norc.org
}
\description{
spschool
}
\details{
Question Listed below are various areas of government spending. Please indicate whether you would like to see more or less government spending in each area. Remember that if you say "much more," it might require a tax increase to pay for it.
D. Education.
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend more
\item \verb{[3]} spend same
\item \verb{[4]} spend less
\item \verb{[5]} spend much less
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4060/vshow}{GSS Data Explorer page for \code{spschool}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend same \tab spend less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 139 \tab 279 \tab 201 \tab 25 \tab 5 \tab 11 \tab 17 \tab 1534 \cr
   1990 \tab 155 \tab 301 \tab 556 \tab 267 \tab 20 \tab 9 \tab 27 \tab 37 \tab 1372 \cr
   1996 \tab 1572 \tab 371 \tab 621 \tab 222 \tab 47 \tab 21 \tab 17 \tab 33 \tab 2904 \cr
   2006 \tab 2992 \tab 617 \tab 631 \tab 203 \tab 50 \tab 10 \tab 7 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 427 \tab 650 \tab 243 \tab 39 \tab 14 \tab 7 \tab 10 \tab 2867 \cr
   Total \tab 7053 \tab 1855 \tab 2737 \tab 1136 \tab 181 \tab 59 \tab 69 \tab 97 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bustax}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spunemp}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
