% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_159.R
\name{spsibs}
\alias{spsibs}
\title{No of siblings spouse has}
\source{
General Social Survey https://gss.norc.org
}
\description{
spsibs
}
\details{
Question How many brothers and sisters does your (husband/wife) have? (Count those born alive, but no longer living, as well as those alive now. Also include stepbrothers and stepsisters, and children adopted by his/her parents.)
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1491/vshow}{GSS Data Explorer page for \code{spsibs}}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 19 \tab 2 \tab 21 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2213 \tab 38 \tab 127 \tab 13 \tab 6 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 145 \tab 1 \tab 135 \tab 93 \tab 61 \tab 43 \tab 41 \tab 21 \tab 21 \tab 5 \tab 17 \tab 2992 \cr
   Total \tab 2213 \tab 38 \tab 127 \tab 13 \tab 6 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 145 \tab 1 \tab 135 \tab 93 \tab 61 \tab 43 \tab 41 \tab 21 \tab 21 \tab 5 \tab 17 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{sphedpre105plus}}
}
\concept{Family Mobility}
\keyword{variable}
