% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_292.R
\name{sptmkid1}
\alias{sptmkid1}
\title{Time spouse spends with child 1}
\source{
General Social Survey https://gss.norc.org
}
\description{
sptmkid1
}
\details{
Question Would you say that the amount of time your (HUSBAND/WIFE) spends with (NAME OF CHILD) is too much, about the right amount, or too little?
A. First listed child?
}
\section{Values}{

\itemize{
\item \code{1} too much
\item \code{2} about the right amount
\item \code{3} too little
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2643/vshow}{GSS Data Explorer page for \code{sptmkid1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too much \tab about the right amount \tab too little \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2474 \tab 8 \tab 178 \tab 133 \tab 3 \tab 21 \tab 2817 \cr
   Total \tab 2474 \tab 8 \tab 178 \tab 133 \tab 3 \tab 21 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}}
}
\concept{Information Society}
\keyword{variable}
