% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_292.R
\name{sptmkid2}
\alias{sptmkid2}
\title{Time spouse spends with child 2}
\source{
General Social Survey https://gss.norc.org
}
\description{
sptmkid2
}
\details{
Question Would you say that the amount of time your (HUSBAND/WIFE) spends with (NAME OF CHILD) is too much, about the right amount, or too little?
B. Second listed child?
}
\section{Values}{

\itemize{
\item \verb{[1]} too much
\item \verb{[2]} about the right amount
\item \verb{[3]} too little
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2644/vshow}{GSS Data Explorer page for \code{sptmkid2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab too much \tab about the right amount \tab too little \tab no answer \tab Total \cr
   2000 \tab 2599 \tab 8 \tab 107 \tab 79 \tab 24 \tab 2817 \cr
   Total \tab 2599 \tab 8 \tab 107 \tab 79 \tab 24 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{sptmkid1}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}}
}
\concept{Information Society}
\keyword{variable}
