% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{spwkbaby}
\alias{spwkbaby}
\title{Did spouse work with young children}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwkbaby
}
\details{
Question What about your spouse/partner at that time -- did he or she work outside the home full-time, part-time, or not at all…
B. And what about when a child was under school age?
}
\section{Values}{

\itemize{
\item \code{1} worked full-time
\item \code{2} worked part-time
\item \code{3} stayed home
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4424/vshow}{GSS Data Explorer page for \code{spwkbaby}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{repairs1}},
\code{\link{shopfood}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
