% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{spwksup}
\alias{spwksup}
\title{Does spouse supervise anyone}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwksup
}
\details{
Question In your (husband/wife/partner)'s job, does (he/she) supervise anyone who is directly responsible to (him/her)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7093/vshow}{GSS Data Explorer page for \code{spwksup}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab 1859 \tab 172 \tab 310 \tab 5 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 3227 \tab 282 \tab 496 \tab 24 \tab - \tab 3 \tab 4032 \cr
   2022 \tab 2876 \tab 265 \tab 381 \tab 9 \tab 12 \tab 1 \tab 3544 \cr
   2024 \tab 2728 \tab 223 \tab 346 \tab 12 \tab - \tab - \tab 3309 \cr
   Total \tab 10690 \tab 942 \tab 1533 \tab 50 \tab 14 \tab 4 \tab 13233 \cr
}
}

\keyword{variable}
