% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{spwrkgvt}
\alias{spwrkgvt}
\title{Spouse working for government}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwrkgvt
}
\details{
Question (Is/Was)(s/he)  employed by the federal, state, or local government or by a private employer (including non-profit organizations)?
}
\section{Values}{

\itemize{
\item \verb{[1]} government
\item \verb{[2]} private employer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4626/vshow}{GSS Data Explorer page for \code{spwrkgvt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab government \tab private employer \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3749 \tab 139 \tab 587 \tab 20 \tab 15 \tab 4510 \cr
   Total \tab 3749 \tab 139 \tab 587 \tab 20 \tab 15 \tab 4510 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{otherwrk}},
\code{\link{skltrain}}
}
\concept{Work Orientation}
\keyword{variable}
