% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{spwrkslf}
\alias{spwrkslf}
\title{Spouse self-emp. or works for somebody}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwrkslf
}
\details{
Question A. What kind of work (does/did) your (SPOUSE) normally do? That is, what (is/was) (his/her) job called?
OCCUPATION: (See REMARKS)
B. What (does/did) (SPOUSE) actually do in that job? Tell me, what (are/were) some of (his/her) main duties?
C. What kind of place (does/did) (SPOUSE) work for?
INDUSTRY: (See REMARKS)
D. What (do/did) they (make/do)?
E. (Is/Was) (he/she) self-employed or (does/did) (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/29/vshow}{GSS Data Explorer page for \code{spwrkslf}}.
}

\seealso{
Other Core: 
\code{\link{cohrs2}},
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{sppres}},
\code{\link{sppres80}}

Other Occupation: 
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}}
}
\concept{Core}
\concept{Job}
\concept{Occupation}
\concept{Work}
\keyword{variable}
