% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{srcheng1}
\alias{srcheng1}
\title{Does r know the name of a search engine}
\source{
General Social Survey https://gss.norc.org
}
\description{
srcheng1
}
\details{
Question Now we'd like to ask you a few questions about terms that people use to talk about the Internet and World Wide Web.
A. Do you know the name of any search engine?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2511/vshow}{GSS Data Explorer page for \code{srcheng1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2075 \tab 526 \tab 140 \tab 76 \tab 2817 \cr
   Total \tab 2075 \tab 526 \tab 140 \tab 76 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng2}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
