% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{srcheng2}
\alias{srcheng2}
\title{Name of search engine given by r}
\source{
General Social Survey https://gss.norc.org
}
\description{
srcheng2
}
\details{
Question Now we'd like to ask you a few questions about terms that people use to talk about the Internet and World Wide Web.
B. Please name one.
}
\section{Values}{

\itemize{
\item \verb{[1]} yahoo
\item \verb{[2]} lycos
\item \verb{[3]} search.com
\item \verb{[4]} gopher
\item \verb{[5]} all the web
\item \verb{[6]} msn
\item \verb{[7]} ebay
\item \verb{[8]} infoseek
\item \verb{[9]} altavista
\item \verb{[10]} explorer
\item \verb{[11]} netscape
\item \verb{[12]} isearch
\item \verb{[13]} excite
\item \verb{[14]} aol
\item \verb{[15]} java
\item \verb{[16]} webcrawler
\item \verb{[17]} hotbot
\item \verb{[18]} google
\item \verb{[19]} copernic
\item \verb{[20]} ask jeeves
\item \verb{[21]} iwon
\item \verb{[22]} dogpile
\item \verb{[23]} streetviolence.com
\item \verb{[24]} browser
\item \verb{[25]} northern lights
\item \verb{[26]} www.irs.com
\item \verb{[27]} netfind
\item \verb{[28]} keysearch
\item \verb{[29]} metacrawler
\item \verb{[30]} uproar.com
\item \verb{[31]} siap
\item \verb{[32]} go
\item \verb{[33]} myway
\item \verb{[34]} fuse
\item \verb{[35]} espn
\item \verb{[36]} uswest
\item \verb{[37]} terra.com
\item \verb{[38]} goto.com
\item \verb{[39]} bizrate
\item \verb{[40]} fastsearch
\item \verb{[41]} sherlock
\item \verb{[42]} lynco
\item \verb{[43]} inference
\item \verb{[44]} 4anything
\item \verb{[45]} navistar
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2512/vshow}{GSS Data Explorer page for \code{srcheng2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab yahoo \tab lycos \tab search.com \tab gopher \tab all the web \tab msn \tab ebay \tab infoseek \tab altavista \tab explorer \tab netscape \tab isearch \tab excite \tab aol \tab java \tab webcrawler \tab hotbot \tab google \tab copernic \tab ask jeeves \tab iwon \tab dogpile \tab streetviolence.com \tab browser \tab northern lights \tab www.irs.com \tab netfind \tab keysearch \tab metacrawler \tab uproar.com \tab siap \tab go \tab myway \tab fuse \tab espn \tab goto.com \tab bizrate \tab sherlock \tab lynco \tab 4anything \tab navistar \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2000 \tab 2215 \tab 315 \tab 33 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 10 \tab 26 \tab 7 \tab 14 \tab 1 \tab 19 \tab 17 \tab 1 \tab 4 \tab 7 \tab 12 \tab 1 \tab 3 \tab 5 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 82 \tab - \tab 2817 \cr
   Total \tab 2215 \tab 315 \tab 33 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 10 \tab 26 \tab 7 \tab 14 \tab 1 \tab 19 \tab 17 \tab 1 \tab 4 \tab 7 \tab 12 \tab 1 \tab 3 \tab 5 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 82 \tab 0 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artgn}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyinfgn}},
\code{\link{buyit12}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{gamegn}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{huclean}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{inperson}},
\code{\link{intaccss}},
\code{\link{invest12}},
\code{\link{investgn}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobhrw}},
\code{\link{jobinfo}},
\code{\link{jobminw}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{letters}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{meetings}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{newjobgn}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numcntct}},
\code{\link{numhome}},
\code{\link{numprobs}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{peoplegn}},
\code{\link{perfrmgn}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polcamp}},
\code{\link{polcangn}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldisgn}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polfrel}},
\code{\link{polinf12}},
\code{\link{polinfgn}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{polpapr}},
\code{\link{poltough}},
\code{\link{poltv}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{travelgn}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}

Other Computers And Internet: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{autonojb}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{intaccss}},
\code{\link{internet}},
\code{\link{intrhome}},
\code{\link{invest12}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numhome}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polinf12}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{poltough}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
