% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{ssfchild}
\alias{ssfchild}
\title{Same sex female couple raise child as well as male-female couple}
\source{
General Social Survey https://gss.norc.org
}
\description{
ssfchild
}
\details{
Question To what extent do you agree or disagree with the following statements?
A. A same sex female couple can bring up a child as well as a male-female couple.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4457/vshow}{GSS Data Explorer page for \code{ssfchild}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 124 \tab 451 \tab 152 \tab 315 \tab 188 \tab 62 \tab 10 \tab - \tab 1974 \cr
   2022 \tab 2381 \tab 295 \tab 404 \tab 181 \tab 138 \tab 75 \tab 54 \tab 9 \tab 7 \tab 3544 \cr
   Total \tab 3053 \tab 419 \tab 855 \tab 333 \tab 453 \tab 263 \tab 116 \tab 19 \tab 7 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
