% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{ssmchild}
\alias{ssmchild}
\title{Same sex male couple raise child as well as male-female couple}
\source{
General Social Survey https://gss.norc.org
}
\description{
ssmchild
}
\details{
Question To what extent do you agree or disagree with the following statements?
B. A same sex male couple can bring up a child as well as a male-female couple.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4458/vshow}{GSS Data Explorer page for \code{ssmchild}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 116 \tab 411 \tab 158 \tab 330 \tab 213 \tab 63 \tab 11 \tab - \tab 1974 \cr
   2022 \tab 2381 \tab 285 \tab 385 \tab 186 \tab 148 \tab 85 \tab 58 \tab 10 \tab 6 \tab 3544 \cr
   Total \tab 3053 \tab 401 \tab 796 \tab 344 \tab 478 \tab 298 \tab 121 \tab 21 \tab 6 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
