% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{stairs}
\alias{stairs}
\title{R limited in climbing stair's}
\source{
General Social Survey https://gss.norc.org
}
\description{
stairs
}
\details{
Question The following questions are about activities you might do during a typical day. Does your health now limit you in these activities?  If so, how much?
B. Climbing several flights of stairs?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, limited a lot
\item \verb{[2]} yes, limited a little
\item \verb{[3]} no, not limited at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2678/vshow}{GSS Data Explorer page for \code{stairs}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, limited a lot \tab yes, limited a little \tab no, not limited at all \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 168 \tab 200 \tab 1019 \tab 32 \tab 2817 \cr
   Total \tab 1398 \tab 168 \tab 200 \tab 1019 \tab 32 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{crelesse}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{limitedp}},
\code{\link{modact}}
}
\concept{Health Status}
\keyword{variable}
