% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{standup}
\alias{standup}
\title{Stand up for your own opinion}
\source{
General Social Survey https://gss.norc.org
}
\description{
standup
}
\details{
Question Some people think that it's important to stand up for your own opinion even if it makes others around you  uncomfortable. Others think that it's better to keep your views to yourself if they would make others around you uncomfortable. Which position comes closer to your view?
}
\section{Values}{

\itemize{
\item \verb{[1]} should stand up regardless
\item \verb{[2]} keep it yourself if offend others
\item \verb{[3]} neither
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1376/vshow}{GSS Data Explorer page for \code{standup}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab should stand up regardless \tab keep it yourself if offend others \tab neither \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1993 \tab - \tab 890 \tab 607 \tab 75 \tab 22 \tab 12 \tab - \tab 1606 \cr
   Total \tab 0 \tab 890 \tab 607 \tab 75 \tab 22 \tab 12 \tab 0 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{bigband}},
\code{\link{blues}},
\code{\link{blugrass}},
\code{\link{camping}},
\code{\link{classicl}},
\code{\link{classics}},
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{conrock}},
\code{\link{country}},
\code{\link{dance}},
\code{\link{dosports}},
\code{\link{drama}},
\code{\link{english}},
\code{\link{excelart}},
\code{\link{folk}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{garden}},
\code{\link{gomusic}},
\code{\link{gospel}},
\code{\link{grtbooks}},
\code{\link{hosthome}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{huntfish}},
\code{\link{hvymetal}},
\code{\link{impcultr}},
\code{\link{impfinan}},
\code{\link{impgod}},
\code{\link{impjob}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{jazz}},
\code{\link{judgeart}},
\code{\link{latin}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{makeart}},
\code{\link{modpaint}},
\code{\link{moodeasy}},
\code{\link{mostimp1}},
\code{\link{mostimp2}},
\code{\link{mostimp3}},
\code{\link{mostimp4}},
\code{\link{musicals}},
\code{\link{newage}},
\code{\link{obtohelp}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{opoutcme}},
\code{\link{pclit}},
\code{\link{perform}},
\code{\link{plymusic}},
\code{\link{popmusic}},
\code{\link{rap}},
\code{\link{readfict}},
\code{\link{reggae}},
\code{\link{richpoor}},
\code{\link{seemovie}},
\code{\link{selfirst}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}},
\code{\link{united}},
\code{\link{usevcr}},
\code{\link{visitart}}

Other Sociability: 
\code{\link{hosthome}},
\code{\link{makefrnd}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socpars}},
\code{\link{socrel}},
\code{\link{socsibs}}
}
\concept{Culture}
\concept{Sociability}
\keyword{variable}
