% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{stateart}
\alias{stateart}
\title{Feel state gov should provide assistance to art}
\source{
General Social Survey https://gss.norc.org
}
\description{
stateart
}
\details{
Question If arts organizations -- such as art museums, dance, operas, theater groups and symphony orchestras --  need financial assistance to operate, Do you feel that the federal government should provide assistance or not?
A. What about state government? Do you feel that it should provide financial assistance to arts organizations?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2135/vshow}{GSS Data Explorer page for \code{stateart}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 769 \tab 502 \tab 156 \tab 18 \tab 2832 \cr
   Total \tab 1387 \tab 769 \tab 502 \tab 156 \tab 18 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{drama}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{volarts}}
}
\concept{Religion and Culture}
\keyword{variable}
