% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_651.R
\name{stayhome}
\alias{stayhome}
\title{Govt should make people stay at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
stayhome
}
\details{
Question (Do you think the United States government should or should not have the right to do the following at times of severe epidemics?) Demand that people stay at home.
}
\section{Values}{

\itemize{
\item \code{1} definitely should have the right
\item \code{2} probably should have the right
\item \code{3} probably should not have the right
\item \code{4} definitely should not have the right
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7841/vshow}{GSS Data Explorer page for \code{stayhome}}.
}

\keyword{variable}
