% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_143.R
\name{stayorg1}
\alias{stayorg1}
\title{Would take any job to keep working for this org.}
\source{
General Social Survey https://gss.norc.org
}
\description{
stayorg1
}
\details{
Question Please tell me how much you agree or disagree with the following statements. Would you say that you strongly agree, agree, disagree, or strongly disagree?
D. I would take almost any job to keep working for this organization.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1343/vshow}{GSS Data Explorer page for \code{stayorg1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 92 \tab 222 \tab 421 \tab 142 \tab 19 \tab 16 \tab 1517 \cr
   Total \tab 605 \tab 92 \tab 222 \tab 421 \tab 142 \tab 19 \tab 16 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{chngeorg}},
\code{\link{helporg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{samevals}},
\code{\link{stayorg2}},
\code{\link{wrkmuch}}
}
\concept{Work Organizations}
\keyword{variable}
