% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_143.R
\name{stayorg2}
\alias{stayorg2}
\title{Turn down another job for pay to stay this org.}
\source{
General Social Survey https://gss.norc.org
}
\description{
stayorg2
}
\details{
Question Please tell me how much you agree or disagree with the following statements. Would you say that you strongly agree, agree, disagree, or strongly disagree?
G. I would turn down another job for more pay in order to stay with this organization.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1346/vshow}{GSS Data Explorer page for \code{stayorg2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 113 \tab 215 \tab 376 \tab 153 \tab 42 \tab 13 \tab 1517 \cr
   Total \tab 605 \tab 113 \tab 215 \tab 376 \tab 153 \tab 42 \tab 13 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{chngeorg}},
\code{\link{helporg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{samevals}},
\code{\link{stayorg1}},
\code{\link{wrkmuch}}
}
\concept{Work Organizations}
\keyword{variable}
