% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_320.R
\name{stockops}
\alias{stockops}
\title{R hold any stock options of r's company}
\source{
General Social Survey https://gss.norc.org
}
\description{
stockops
}
\details{
Question Do you currently hold any stock options in your company (vested or unvested)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2854/vshow}{GSS Data Explorer page for \code{stockops}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{ownstock}},
\code{\link{stockval}},
\code{\link{workfor}},
\code{\link{workfor1}}

Other Job: 
\code{\link{ownstock}},
\code{\link{stockval}}
}
\concept{Employee Compensation}
\concept{Job}
\keyword{variable}
