% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{stockoptyr}
\alias{stockoptyr}
\title{Granted stock options in past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
stockoptyr
}
\details{
Question Were you granted stock options in the company where you work in 2021? A stock option is a right to purchase stock at a set price for several years into the future, typically ten years. For example, one stock option at $10 means that you can purchase yo
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7848/vshow}{GSS Data Explorer page for \code{stockoptyr}}.
}

\keyword{variable}
