% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_651.R
\name{stockval1}
\alias{stockval1}
\title{Estimate of total value of stocks if sold today}
\source{
General Social Survey https://gss.norc.org
}
\description{
stockval1
}
\details{
Question Please give a general estimate of how much cash you would get if all this stock were sold today.
}
\section{Values}{

\itemize{
\item \verb{[1]} less than $499
\item \verb{[2]} $500 to $999
\item \verb{[3]} $1,000 to $2,999
\item \verb{[4]} $3,000 to $4,999
\item \verb{[5]} $5,000 to $9,999
\item \verb{[6]} $10,000 to $19,999
\item \verb{[7]} $20,000 to $39,999
\item \verb{[8]} $40,000 to $79,999
\item \verb{[9]} $80,000 to $199,999
\item \verb{[10]} $200,000 to $499,999
\item \verb{[11]} $500,000 to $999,000
\item \verb{[12]} over $1 million
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7845/vshow}{GSS Data Explorer page for \code{stockval1}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab less than $499 \tab $500 to $999 \tab $1,000 to $2,999 \tab $3,000 to $4,999 \tab $5,000 to $9,999 \tab $10,000 to $19,999 \tab $20,000 to $39,999 \tab $40,000 to $79,999 \tab $80,000 to $199,999 \tab $200,000 to $499,999 \tab $500,000 to $999,000 \tab over $1 million \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3423 \tab 14 \tab 4 \tab 4 \tab 13 \tab 11 \tab 9 \tab 8 \tab 10 \tab 5 \tab 10 \tab 2 \tab 5 \tab 24 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 3423 \tab 14 \tab 4 \tab 4 \tab 13 \tab 11 \tab 9 \tab 8 \tab 10 \tab 5 \tab 10 \tab 2 \tab 5 \tab 24 \tab 1 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
