% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{stoprndm}
\alias{stoprndm}
\title{Should authorities have right to search ppl in the street}
\source{
General Social Survey https://gss.norc.org
}
\description{
stoprndm
}
\details{
Question Suppose the government suspected that a terrorist act was about to happen. Do you think the authorities should have the right to…
C. Stop and search people in the street at random?
}
\section{Values}{

\itemize{
\item \code{1} definitely should have right
\item \code{2} probably should have right
\item \code{3} probably should not have right
\item \code{4} definitely should not have right
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4116/vshow}{GSS Data Explorer page for \code{stoprndm}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{polsfair}},
\code{\link{tapphone}},
\code{\link{wotrial}},
\code{\link{youinflu}}

Other United States: 
\code{\link{tapphone}},
\code{\link{wotrial}}
}
\concept{ISSP Role of Gov't}
\concept{United States}
\keyword{variable}
