% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{stormtxt}
\alias{stormtxt}
\title{Do you see lightning before hearing thunder}
\source{
General Social Survey https://gss.norc.org
}
\description{
stormtxt
}
\details{
Question Lightning and thunder happen at the same time, but you see the lightning before you hear the thunder. Explain why this is so.
}
\section{Values}{

\itemize{
\item \code{1} complete
\item \code{2} partial
\item \code{3} unsatisfactory/incorrect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3474/vshow}{GSS Data Explorer page for \code{stormtxt}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
