% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{stress}
\alias{stress}
\title{How often does r find work stressful}
\source{
General Social Survey https://gss.norc.org
}
\description{
stress
}
\details{
Question Now some more questions about your working conditions. one code for each item below to show how often it applies to your work. How often...
C. Do you find your work stressful?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1989, 1998, 2002, 2004, 2010, 2012, 2014, 2018, 2022 \cr
   A/B/C/D \tab 2006 \cr
   B/C/-/- \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3012/vshow}{GSS Data Explorer page for \code{stress}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not available in this year \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1989 \tab - \tab - \tab 82 \tab 261 \tab 382 \tab 109 \tab 37 \tab 3 \tab 663 \tab - \tab 1537 \cr
   1998 \tab - \tab - \tab 89 \tab 229 \tab 378 \tab 91 \tab 42 \tab 1 \tab 2002 \tab - \tab 2832 \cr
   2002 \tab - \tab - \tab 161 \tab 391 \tab 756 \tab 329 \tab 137 \tab 6 \tab 985 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab 213 \tab 475 \tab 797 \tab 242 \tab 95 \tab 5 \tab 985 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab 353 \tab 595 \tab 1154 \tab 449 \tab 185 \tab 3 \tab 1771 \tab - \tab 4510 \cr
   2010 \tab - \tab - \tab 110 \tab 284 \tab 475 \tab 215 \tab 78 \tab 1 \tab 881 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 118 \tab 320 \tab 478 \tab 162 \tab 66 \tab 5 \tab 825 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 113 \tab 247 \tab 559 \tab 225 \tab 99 \tab 1 \tab 1294 \tab - \tab 2538 \cr
   2016 \tab 911 \tab - \tab 96 \tab 207 \tab 467 \tab 130 \tab 32 \tab 2 \tab 1022 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab 107 \tab 304 \tab 628 \tab 264 \tab 110 \tab 3 \tab 932 \tab - \tab 2348 \cr
   2022 \tab 1557 \tab - \tab 171 \tab 409 \tab 906 \tab 339 \tab 128 \tab 5 \tab 14 \tab 15 \tab 3544 \cr
   Total \tab 2468 \tab 0 \tab 1613 \tab 3722 \tab 6980 \tab 2555 \tab 1009 \tab 35 \tab 11374 \tab 15 \tab 29771 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
