% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{stresses}
\alias{stresses}
\title{How likely x s situation caused by stress}
\source{
General Social Survey https://gss.norc.org
}
\description{
stresses
}
\details{
Question In your opinion, how likely is it that NAME's situation might be caused by (READ a-f) -- very likely, somewhat likely, not very likely, or not at all likely?
D. Stressful circumstances in his/her life
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1991/vshow}{GSS Data Explorer page for \code{stresses}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 523 \tab 706 \tab 102 \tab 57 \tab 40 \tab 16 \tab 2904 \cr
   2006 \tab 3073 \tab 540 \tab 732 \tab 84 \tab 36 \tab 31 \tab 14 \tab 4510 \cr
   2018 \tab 1175 \tab 387 \tab 580 \tab 102 \tab 59 \tab 16 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 1450 \tab 2018 \tab 288 \tab 152 \tab 87 \tab 59 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
