% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{stresswk}
\alias{stresswk}
\title{Stress in past 2 wks}
\source{
General Social Survey https://gss.norc.org
}
\description{
stresswk
}
\details{
Question During the past two weeks, would you say that you experienced a lot of stress, a moderate amount of stress, relatively little stress, or almost no stress at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} a lot of stress
\item \verb{[2]} a moderate amount of stress
\item \verb{[3]} relatively little stress
\item \verb{[4]} almost no stress at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6196/vshow}{GSS Data Explorer page for \code{stresswk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a lot of stress \tab a moderate amount of stress \tab relatively little stress \tab almost no stress at all \tab no answer \tab Total \cr
   2016 \tab 2389 \tab 110 \tab 171 \tab 129 \tab 65 \tab 3 \tab 2867 \cr
   Total \tab 2389 \tab 110 \tab 171 \tab 129 \tab 65 \tab 3 \tab 2867 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{rushed}}
}
\concept{Core}
\keyword{variable}
