% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{strike}
\alias{strike}
\title{Picketing for labor strike}
\source{
General Social Survey https://gss.norc.org
}
\description{
strike
}
\details{
Question Have you ever taken part in:
A. Picketing for a labor strike?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/653/vshow}{GSS Data Explorer page for \code{strike}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   1973 \tab - \tab 143 \tab 1361 \tab 1504 \cr
   Total \tab 0 \tab 143 \tab 1361 \tab 1504 \cr
}
}

\seealso{
Other Core: 
\code{\link{antiwar}},
\code{\link{civright}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitnum}},
\code{\link{prowar}},
\code{\link{school}}
}
\concept{Core}
\keyword{variable}
