% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{strmgtsup}
\alias{strmgtsup}
\title{Mangement supports stress prevention}
\source{
General Social Survey https://gss.norc.org
}
\description{
strmgtsup
}
\details{
Question To what extent to do you agree or disagree with the following statements? Senior management show support for stress prevention through involvement and commitment.
}
\section{Values}{

\itemize{
\item \code{1} strongly disagree
\item \code{2} disagree
\item \code{3} neither agree nor disagree
\item \code{4} agree
\item \code{5} strongly agree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7856/vshow}{GSS Data Explorer page for \code{strmgtsup}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly disagree \tab disagree \tab neither agree nor disagree \tab agree \tab strongly agree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1548 \tab 263 \tab 280 \tab 566 \tab 622 \tab 188 \tab 30 \tab 24 \tab 23 \tab 3544 \cr
   Total \tab 1548 \tab 263 \tab 280 \tab 566 \tab 622 \tab 188 \tab 30 \tab 24 \tab 23 \tab 3544 \cr
}
}

\keyword{variable}
