% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{strngun}
\alias{strngun}
\title{Strong unions needed to protect workers}
\source{
General Social Survey https://gss.norc.org
}
\description{
strngun
}
\details{
Question How much do you agree or disagree with these two statements?
B. Workers need strong trade unions to protect their interests.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
   B/C/- \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4506/vshow}{GSS Data Explorer page for \code{strngun}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 160 \tab 364 \tab 341 \tab 342 \tab 164 \tab 65 \tab 17 \tab 1537 \cr
   2016 \tab 1390 \tab 285 \tab 441 \tab 398 \tab 234 \tab 94 \tab 18 \tab 7 \tab 2867 \cr
   Total \tab 1474 \tab 445 \tab 805 \tab 739 \tab 576 \tab 258 \tab 83 \tab 24 \tab 4404 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{bosswrks}},
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{hiinc}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}}
}
\concept{Work Orientation}
\keyword{variable}
