% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{strredpg}
\alias{strredpg}
\title{Access to stress management}
\source{
General Social Survey https://gss.norc.org
}
\description{
strredpg
}
\details{
Question Do you have access to stress management or stress reduction programs at your current workplace?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2835/vshow}{GSS Data Explorer page for \code{strredpg}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 857 \tab 404 \tab 712 \tab 48 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 478 \tab 719 \tab 46 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 511 \tab 847 \tab 57 \tab 33 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 879 \tab 1002 \tab 68 \tab 18 \tab 22 \tab 3544 \cr
   Total \tab 4600 \tab 2272 \tab 3280 \tab 219 \tab 81 \tab 22 \tab 10474 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{backpain}},
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{usedup}}
}
\concept{Quality of Working Life}
\keyword{variable}
