% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_55.R
\name{studious}
\alias{studious}
\title{Good student}
\source{
General Social Survey https://gss.norc.org
}
\description{
studious
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
13. That he is a good student
}
\section{Values}{

\itemize{
\item \code{1} 1 most desirable
\item \code{2} 3 most desirable
\item \code{3} not mentioned
\item \code{4} 3 least desirable
\item \code{5} 1 least desirable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1975, 1976, 1978, 1980, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/486/vshow}{GSS Data Explorer page for \code{studious}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 15 \tab 62 \tab 822 \tab 444 \tab 157 \tab 4 \tab - \tab 1504 \cr
   1975 \tab - \tab 13 \tab 55 \tab 821 \tab 462 \tab 130 \tab 9 \tab - \tab 1490 \cr
   1976 \tab - \tab 6 \tab 67 \tab 850 \tab 447 \tab 120 \tab 9 \tab - \tab 1499 \cr
   1978 \tab - \tab 7 \tab 67 \tab 830 \tab 496 \tab 120 \tab 10 \tab 2 \tab 1532 \cr
   1980 \tab 975 \tab 2 \tab 25 \tab 283 \tab 147 \tab 32 \tab 3 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 13 \tab 82 \tab 891 \tab 473 \tab 120 \tab 20 \tab - \tab 1599 \cr
   Total \tab 975 \tab 56 \tab 358 \tab 4497 \tab 2469 \tab 679 \tab 55 \tab 3 \tab 9092 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicably}},
\code{\link{cleany}},
\code{\link{controly}},
\code{\link{honesty}},
\code{\link{interest}},
\code{\link{judgmeny}},
\code{\link{mannersy}},
\code{\link{roley}},
\code{\link{successy}}
}
\concept{Core}
\keyword{variable}
