% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_56.R
\name{studiouy}
\alias{studiouy}
\title{Good student-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
studiouy
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
13. That a child is a good student
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 most desirable
\item \verb{[2]} 3 most desirable
\item \verb{[3]} not mentioned
\item \verb{[4]} 3 least desirable
\item \verb{[5]} 1 least desirable
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980, 1984, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/499/vshow}{GSS Data Explorer page for \code{studiouy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab Total \cr
   1980 \tab 976 \tab 4 \tab 26 \tab 296 \tab 128 \tab 33 \tab 5 \tab 1468 \cr
   1984 \tab - \tab 27 \tab 117 \tab 866 \tab 355 \tab 73 \tab 35 \tab 1473 \cr
   1986 \tab 740 \tab 14 \tab 118 \tab 406 \tab 151 \tab 27 \tab 14 \tab 1470 \cr
   Total \tab 1716 \tab 45 \tab 261 \tab 1568 \tab 634 \tab 133 \tab 54 \tab 4411 \cr
}
}

\seealso{
Other Core: 
\code{\link{cleanz}},
\code{\link{considey}},
\code{\link{honestz}},
\code{\link{interesy}},
\code{\link{judgmenz}},
\code{\link{mannersz}},
\code{\link{obeysy}},
\code{\link{responsy}},
\code{\link{successz}}
}
\concept{Core}
\keyword{variable}
