% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{subsamprate}
\alias{subsamprate}
\title{Subsampling rate corresponding to two-phase design}
\source{
General Social Survey https://gss.norc.org
}
\description{
subsamprate
}
\details{
Question Subsampling rate corresponding to two-phase design
}
\section{Values}{

\itemize{
\item \verb{[1]} 100pct subsampling rate
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8042/vshow}{GSS Data Explorer page for \code{subsamprate}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab 0.5 \tab iap \tab 0.2 \tab 100pct subsampling rate \tab 0.3 \tab 0.4 \tab Total \cr
   2004 \tab - \tab 448 \tab 2364 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 874 \tab 3636 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 411 \tab 1612 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 386 \tab 1658 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 488 \tab 1486 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 807 \tab 1731 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 573 \tab 2294 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 248 \tab 2100 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab - \tab - \tab 1326 \tab 744 \tab 1474 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab 532 \tab 1511 \tab 577 \tab 689 \tab 3309 \cr
   Total \tab 0 \tab 4235 \tab 18207 \tab 1276 \tab 2985 \tab 577 \tab 689 \tab 27969 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{batch}}
}
\concept{Sampling and Weights}
\keyword{variable}
