% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_55.R
\name{successy}
\alias{successy}
\title{Tries hard to succeed-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
successy
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
2. That a child tries hard to succeed
}
\section{Values}{

\itemize{
\item \code{1} 1 most desirable
\item \code{2} 3 most desirable
\item \code{3} not mentioned
\item \code{4} 3 least desirable
\item \code{5} 1 least desirable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980, 1984, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/488/vshow}{GSS Data Explorer page for \code{successy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 1 most desirable \tab 3 most desirable \tab not mentioned \tab 3 least desirable \tab 1 least desirable \tab no answer \tab Total \cr
   1980 \tab 976 \tab 20 \tab 71 \tab 295 \tab 89 \tab 12 \tab 5 \tab 1468 \cr
   1984 \tab - \tab 53 \tab 204 \tab 835 \tab 289 \tab 57 \tab 35 \tab 1473 \cr
   1986 \tab 740 \tab 36 \tab 86 \tab 425 \tab 144 \tab 25 \tab 14 \tab 1470 \cr
   Total \tab 1716 \tab 109 \tab 361 \tab 1555 \tab 522 \tab 94 \tab 54 \tab 4411 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicably}},
\code{\link{cleany}},
\code{\link{controly}},
\code{\link{honesty}},
\code{\link{interest}},
\code{\link{judgmeny}},
\code{\link{mannersy}},
\code{\link{roley}},
\code{\link{studious}}
}
\concept{Core}
\keyword{variable}
