% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_331.R
\name{sufadult}
\alias{sufadult}
\title{Would child suffer as adult if other learn}
\source{
General Social Survey https://gss.norc.org
}
\description{
sufadult
}
\details{
Question That was my last question about (John/Mary). Now, I have some questions about children, but not about
(John/Mary). Please tell me whether you Strongly Agree, Agree Somewhat, Disagree Somewhat or Strongly disagree with the following statements:
B. A child would suffer as an adult if others learn that he or she received mental health treatment when young.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree somewhat
\item \verb{[3]} disagree somewhat
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2958/vshow}{GSS Data Explorer page for \code{sufadult}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree somewhat \tab disagree somewhat \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 121 \tab 442 \tab 460 \tab 316 \tab 43 \tab 11 \tab 2765 \cr
   Total \tab 1372 \tab 121 \tab 442 \tab 460 \tab 316 \tab 43 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{comknows}},
\code{\link{failure}},
\code{\link{illphys}},
\code{\link{medkdneg}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{putsoff}},
\code{\link{violpeop}},
\code{\link{violself}}
}
\concept{Mental Health II}
\keyword{variable}
