% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suiage2}
\alias{suiage2}
\title{Age of suicide victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiage2
}
\details{
Question We would like to know a few other things about that person.
B. How old was that person?  Was (he/she)…
}
\section{Values}{

\itemize{
\item \verb{[1]} 10 years or under
\item \verb{[2]} 11-20
\item \verb{[3]} 21-40
\item \verb{[4]} 41 year or older
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5033/vshow}{GSS Data Explorer page for \code{suiage2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 11-20 \tab 21-40 \tab 41 year or older \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab 1347 \tab 6 \tab 11 \tab 5 \tab 1 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 1477 \tab 7 \tab 16 \tab 14 \tab - \tab 3 \tab - \tab 1517 \cr
   Total \tab 2824 \tab 13 \tab 27 \tab 19 \tab 1 \tab 5 \tab 0 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsage3}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsded3}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsknow}},
\code{\link{aidsmar}},
\code{\link{aidsrac2}},
\code{\link{aidsrac3}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssch}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidssex3}},
\code{\link{aidssxed}},
\code{\link{aidswho}},
\code{\link{aidswho2}},
\code{\link{aidswho3}},
\code{\link{cideage}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{cideknew}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidesex3}},
\code{\link{cidewho}},
\code{\link{cidewho2}},
\code{\link{cidewho3}},
\code{\link{suiage}},
\code{\link{suiage3}},
\code{\link{suiknew}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Suicide: 
\code{\link{suiage}},
\code{\link{suiage3}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suiknew}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg1}},
\code{\link{suineg2}},
\code{\link{suinum}},
\code{\link{suinum1}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suirel1}},
\code{\link{suirel2}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\concept{Victims & AIDS}
\keyword{variable}
