% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suicide1}
\alias{suicide1}
\title{Suicide if incurable disease}
\source{
General Social Survey https://gss.norc.org
}
\description{
suicide1
}
\details{
Question Do you think a person has the right to end his or her own life if this person . . .
READ EACH STATEMENT AND CIRCLE ONE CODE FOR EACH.
A. Has an incurable disease?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1977, 1978, 1982, 1983, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/649/vshow}{GSS Data Explorer page for \code{suicide1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1977 \tab - \tab 578 \tab 897 \tab 49 \tab 6 \tab - \tab 1530 \cr
   1978 \tab - \tab 588 \tab 895 \tab 49 \tab - \tab - \tab 1532 \cr
   1982 \tab - \tab 754 \tab 1014 \tab 82 \tab 10 \tab - \tab 1860 \cr
   1983 \tab - \tab 760 \tab 770 \tab 61 \tab 8 \tab - \tab 1599 \cr
   1985 \tab - \tab 675 \tab 805 \tab 52 \tab 2 \tab - \tab 1534 \cr
   1986 \tab - \tab 758 \tab 662 \tab 47 \tab 3 \tab - \tab 1470 \cr
   1988 \tab 493 \tab 492 \tab 454 \tab 37 \tab 5 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 469 \tab 488 \tab 47 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 519 \tab 354 \tab 54 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 579 \tab 408 \tab 35 \tab 2 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 611 \tab 416 \tab 50 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 1211 \tab 663 \tab 95 \tab 8 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 1197 \tab 662 \tab 97 \tab 4 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 1132 \tab 644 \tab 91 \tab 4 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 1025 \tab 745 \tab 101 \tab 6 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 524 \tab 364 \tab 19 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 535 \tab 340 \tab 26 \tab 5 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 1169 \tab 754 \tab 62 \tab 7 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 772 \tab 509 \tab 43 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 834 \tab 543 \tab 46 \tab 7 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 742 \tab 515 \tab 41 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 983 \tab 649 \tab 39 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 1198 \tab 630 \tab 44 \tab 16 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 976 \tab 521 \tab 49 \tab 13 \tab - \tab 2348 \cr
   2021 \tab 2552 \tab 887 \tab 325 \tab 260 \tab 4 \tab 4 \tab 4032 \cr
   2022 \tab 2373 \tab 803 \tab 324 \tab 34 \tab 6 \tab 4 \tab 3544 \cr
   Total \tab 22164 \tab 20771 \tab 15351 \tab 1610 \tab 136 \tab 8 \tab 60040 \cr
}
}

\seealso{
Other Core: 
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}}
}
\concept{Core}
\keyword{variable}
