% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suicide3}
\alias{suicide3}
\title{Suicide if dishonored family}
\source{
General Social Survey https://gss.norc.org
}
\description{
suicide3
}
\details{
Question Do you think a person has the right to end his or her own life if this person . . .
READ EACH STATEMENT AND CIRCLE ONE CODE FOR EACH.
C. Has dishonored his or her family?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1977, 1978, 1982, 1983, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/651/vshow}{GSS Data Explorer page for \code{suicide3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1977 \tab - \tab 114 \tab 1391 \tab 18 \tab 7 \tab - \tab 1530 \cr
   1978 \tab - \tab 96 \tab 1428 \tab 7 \tab 1 \tab - \tab 1532 \cr
   1982 \tab - \tab 142 \tab 1674 \tab 32 \tab 12 \tab - \tab 1860 \cr
   1983 \tab - \tab 107 \tab 1459 \tab 24 \tab 9 \tab - \tab 1599 \cr
   1985 \tab - \tab 121 \tab 1373 \tab 36 \tab 4 \tab - \tab 1534 \cr
   1986 \tab - \tab 87 \tab 1361 \tab 19 \tab 3 \tab - \tab 1470 \cr
   1988 \tab 493 \tab 66 \tab 898 \tab 18 \tab 6 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 70 \tab 913 \tab 17 \tab 6 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 74 \tab 823 \tab 28 \tab 3 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 62 \tab 939 \tab 20 \tab 3 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 81 \tab 972 \tab 24 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 187 \tab 1745 \tab 36 \tab 9 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 194 \tab 1728 \tab 34 \tab 4 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 180 \tab 1645 \tab 40 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 153 \tab 1681 \tab 38 \tab 5 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 79 \tab 819 \tab 10 \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 109 \tab 786 \tab 8 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 172 \tab 1792 \tab 24 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 127 \tab 1182 \tab 17 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 153 \tab 1250 \tab 24 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 140 \tab 1131 \tab 25 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 197 \tab 1442 \tab 32 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 193 \tab 1671 \tab 17 \tab 7 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 194 \tab 1345 \tab 16 \tab 4 \tab - \tab 2348 \cr
   2021 \tab 2556 \tab 145 \tab 1172 \tab 151 \tab 1 \tab 7 \tab 4032 \cr
   2022 \tab 2373 \tab 168 \tab 974 \tab 20 \tab 1 \tab 8 \tab 3544 \cr
   Total \tab 22168 \tab 3411 \tab 33594 \tab 735 \tab 117 \tab 15 \tab 60040 \cr
}
}

\seealso{
Other Core: 
\code{\link{letdie1}},
\code{\link{letdie2}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide4}}
}
\concept{Core}
\keyword{variable}
