% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suiclse2}
\alias{suiclse2}
\title{How close wereyou to 2nd closest suicide victim}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiclse2
}
\details{
Question On a scale of 1 to 5 how close did you feel to that person at the time of his or her death? Use 1 for not close at all and 5 for very close.                         
(HAND CARD I6)
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - not close at all
\item \verb{[5]} 5 - very close
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6191/vshow}{GSS Data Explorer page for \code{suiclse2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 1 - not close at all \tab 2 \tab 3 \tab 4 \tab 5 - very close \tab no answer \tab Total \cr
   2016 \tab 2425 \tab 125 \tab 97 \tab 110 \tab 47 \tab 39 \tab 24 \tab 2867 \cr
   Total \tab 2425 \tab 125 \tab 97 \tab 110 \tab 47 \tab 39 \tab 24 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suirel2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
