% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{suiknew}
\alias{suiknew}
\title{#People r knows suicide in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiknew
}
\details{
Question Within the past 12 months, how many people have you known personally that have committed suicide?
}
\section{Values}{

\itemize{
\item \code{7} seven or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5025/vshow}{GSS Data Explorer page for \code{suiknew}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab no answer \tab 4 \tab Total \cr
   1990 \tab - \tab 1238 \tab 109 \tab 17 \tab 6 \tab 2 \tab - \tab 1372 \cr
   1991 \tab - \tab 1344 \tab 133 \tab 25 \tab 11 \tab 3 \tab 1 \tab 1517 \cr
   Total \tab 0 \tab 2582 \tab 242 \tab 42 \tab 17 \tab 5 \tab 1 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
