% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suimths1}
\alias{suimths1}
\title{How many months since closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suimths1
}
\details{
Question MONTHS:
}
\section{Values}{

\itemize{
\item \verb{[0]} less than one
\item \verb{[97]} no month given
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6187/vshow}{GSS Data Explorer page for \code{suimths1}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab less than one \tab 1 \tab 10 \tab 11 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no month given \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab 2093 \tab 221 \tab 14 \tab 4 \tab 9 \tab 19 \tab 6 \tab 11 \tab 11 \tab 30 \tab 5 \tab 3 \tab 5 \tab 395 \tab 8 \tab 33 \tab - \tab 2867 \cr
   Total \tab 2093 \tab 221 \tab 14 \tab 4 \tab 9 \tab 19 \tab 6 \tab 11 \tab 11 \tab 30 \tab 5 \tab 3 \tab 5 \tab 395 \tab 8 \tab 33 \tab 0 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suiage3}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}},
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suiknew}},
\code{\link{suimths2}},
\code{\link{suineg1}},
\code{\link{suineg2}},
\code{\link{suinum}},
\code{\link{suinum1}},
\code{\link{suirac2}},
\code{\link{suirac3}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suireg3}},
\code{\link{suirel1}},
\code{\link{suirel2}},
\code{\link{suisex}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiwho}},
\code{\link{suiwho2}},
\code{\link{suiwho3}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
