% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suireg}
\alias{suireg}
\title{Region of suicide victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suireg
}
\details{
Question We would like to know a few other things about that person.
D. What state did this person live in?
}
\section{Values}{

\itemize{
\item \verb{[0]} foreign
\item \verb{[1]} new england
\item \verb{[2]} middle atlantic
\item \verb{[3]} e. nor. central
\item \verb{[4]} w. nor. central
\item \verb{[5]} south atlantic
\item \verb{[6]} e. sou. central
\item \verb{[7]} w. sou. central
\item \verb{[8]} mountain
\item \verb{[9]} pacific
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5030/vshow}{GSS Data Explorer page for \code{suireg}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab foreign \tab new england \tab middle atlantic \tab e. nor. central \tab w. nor. central \tab south atlantic \tab e. sou. central \tab w. sou. central \tab mountain \tab pacific \tab no answer \tab Total \cr
   1990 \tab 1238 \tab 2 \tab 10 \tab 18 \tab 20 \tab 11 \tab 14 \tab 14 \tab 15 \tab 10 \tab 17 \tab 3 \tab 1372 \cr
   1991 \tab 1344 \tab 3 \tab 5 \tab 19 \tab 34 \tab 14 \tab 28 \tab 15 \tab 10 \tab 12 \tab 30 \tab 3 \tab 1517 \cr
   Total \tab 2582 \tab 5 \tab 15 \tab 37 \tab 54 \tab 25 \tab 42 \tab 29 \tab 25 \tab 22 \tab 47 \tab 6 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suirac2}},
\code{\link{suirace}},
\code{\link{suireg2}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
