% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suirel2}
\alias{suirel2}
\title{R's relationship to 2nd closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suirel2
}
\details{
Question Please tell me the relationship category of the person you knew second best that died by suicide.
(HAND CARD I4)
}
\section{Values}{

\itemize{
\item \verb{[1]} husband or wife
\item \verb{[2]} partner or lover
\item \verb{[3]} son or daughter
\item \verb{[4]} mother or father
\item \verb{[5]} brother or sister
\item \verb{[6]} step-brother or step-sister
\item \verb{[7]} step-parent
\item \verb{[8]} step-son or step-daughter
\item \verb{[9]} cousin
\item \verb{[10]} other relative
\item \verb{[11]} friend
\item \verb{[12]} neighbor
\item \verb{[13]} co-worker
\item \verb{[14]} patient
\item \verb{[15]} acquaintance
\item \verb{[16]} ex-spouse or ex-partner
\item \verb{[18]} classmate
\item \verb{[19]} relative of a friend
\item \verb{[20]} student
\item \verb{[21]} friend of a friend
\item \verb{[30]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6189/vshow}{GSS Data Explorer page for \code{suirel2}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab husband or wife \tab partner or lover \tab son or daughter \tab mother or father \tab step-parent \tab cousin \tab other relative \tab friend \tab neighbor \tab co-worker \tab patient \tab acquaintance \tab ex-spouse or ex-partner \tab classmate \tab relative of a friend \tab student \tab friend of a friend \tab other \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2425 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 28 \tab 64 \tab 161 \tab 15 \tab 22 \tab 6 \tab 84 \tab 2 \tab 2 \tab 3 \tab 4 \tab 6 \tab 11 \tab 1 \tab 24 \tab 2867 \cr
   Total \tab 2425 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 28 \tab 64 \tab 161 \tab 15 \tab 22 \tab 6 \tab 84 \tab 2 \tab 2 \tab 3 \tab 4 \tab 6 \tab 11 \tab 1 \tab 24 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
