% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suisex3}
\alias{suisex3}
\title{Sex of suicidevictim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suisex3
}
\details{
Question We would like to know a few other things about that person.
A. Was that person male or female?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5037/vshow}{GSS Data Explorer page for \code{suisex3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1990 \tab 1364 \tab 5 \tab 1 \tab 2 \tab 1372 \cr
   1991 \tab 1502 \tab 8 \tab 4 \tab 3 \tab 1517 \cr
   Total \tab 2866 \tab 13 \tab 5 \tab 5 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suirac2}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suisex2}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
