% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suitime1}
\alias{suitime1}
\title{How many years and months since closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suitime1
}
\details{
Question How many months and years has it been since that person's death?
IF RESPONDENT PROVIDES AT LEAST A YEAR AMOUNT, PLEASE SELECT THE FIRST OPTION.
}
\section{Values}{

\itemize{
\item \verb{[1]} respondent provides number of months or years
\item \verb{[2]} respondent provides verbatim response uncodeable in months/years
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6186/vshow}{GSS Data Explorer page for \code{suitime1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab respondent provides number of months or years \tab respondent provides verbatim response uncodeable in months/years \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2093 \tab 700 \tab 42 \tab 7 \tab 25 \tab 2867 \cr
   Total \tab 2093 \tab 700 \tab 42 \tab 7 \tab 25 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suirel2}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
