% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suiwho2}
\alias{suiwho2}
\title{R's relationship to 2nd closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiwho2
}
\details{
Question Think about the person you have known next best, who has committed suicide. Please tell me the letter of the category on the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \code{1} husband or wife
\item \code{2} partner or lover
\item \code{3} son or daughter
\item \code{4} other relative
\item \code{5} friend
\item \code{6} neighbor
\item \code{7} co-worker
\item \code{8} acquaintance
\item \code{9} patient
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5031/vshow}{GSS Data Explorer page for \code{suiwho2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab other relative \tab friend \tab neighbor \tab co-worker \tab acquaintance \tab no answer \tab patient \tab Total \cr
   1990 \tab 1347 \tab 2 \tab 7 \tab 4 \tab 2 \tab 8 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 1477 \tab - \tab 15 \tab 1 \tab 4 \tab 15 \tab 3 \tab 2 \tab 1517 \cr
   Total \tab 2824 \tab 2 \tab 22 \tab 5 \tab 6 \tab 23 \tab 5 \tab 2 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suirac2}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
