% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{sunsch16}
\alias{sunsch16}
\title{How frequently r attended sunday school}
\source{
General Social Survey https://gss.norc.org
}
\description{
sunsch16
}
\details{
Question When you were growing up, did you attend Sunday school or religious instruction classes regularly, most of the time, some of the time, or never?
}
\section{Values}{

\itemize{
\item \verb{[1]} regularly
\item \verb{[2]} most of the time
\item \verb{[3]} some of the time
\item \verb{[4]} never
\item \verb{[5]} att parochial sch
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1069/vshow}{GSS Data Explorer page for \code{sunsch16}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab regularly \tab most of the time \tab some of the time \tab never \tab att parochial sch \tab no answer \tab Total \cr
   1988 \tab - \tab 763 \tab 247 \tab 305 \tab 113 \tab 47 \tab 6 \tab 1481 \cr
   1998 \tab 1387 \tab 781 \tab 186 \tab 284 \tab 137 \tab 48 \tab 9 \tab 2832 \cr
   Total \tab 1387 \tab 1544 \tab 433 \tab 589 \tab 250 \tab 95 \tab 15 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}}
}
\concept{Religion}
\keyword{variable}
