% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{supcares}
\alias{supcares}
\title{Supervisor concerned about welfare}
\source{
General Social Survey https://gss.norc.org
}
\description{
supcares
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
G. My supervisor is concerned with the welfare of those under him or her
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2012, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2802/vshow}{GSS Data Explorer page for \code{supcares}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 885 \tab 563 \tab 148 \tab 134 \tab 39 \tab 27 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 809 \tab 550 \tab 149 \tab 123 \tab 70 \tab 33 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 547 \tab 374 \tab 108 \tab 90 \tab 29 \tab 39 \tab - \tab 2044 \cr
   2012 \tab 795 \tab 502 \tab 399 \tab 88 \tab 105 \tab 46 \tab 39 \tab - \tab 1974 \cr
   2014 \tab 1288 \tab 627 \tab 392 \tab 123 \tab 68 \tab 25 \tab 15 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 715 \tab 461 \tab 137 \tab 70 \tab 27 \tab 38 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 957 \tab 632 \tab 183 \tab 124 \tab 42 \tab 25 \tab 26 \tab 3544 \cr
   Total \tab 9140 \tab 5042 \tab 3371 \tab 936 \tab 714 \tab 278 \tab 216 \tab 26 \tab 19723 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{cowrkint}},
\code{\link{fringeok}},
\code{\link{haveinfo}},
\code{\link{hlpequip}},
\code{\link{jobsecok}},
\code{\link{opdevel}},
\code{\link{promtefr}},
\code{\link{wkfreedm}}

Other Workplace Conflict: 
\code{\link{jobsecok}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
