% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{supduty1}
\alias{supduty1}
\title{You train employees safe use equipment}
\source{
General Social Survey https://gss.norc.org
}
\description{
supduty1
}
\details{
Question Here is a card listing 7 different kinds of oversight activities. Which ones do you do as part of your supervisory responsibilities?  Please mention all that apply:
A. Train employees in the safe use of equipment or tools.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1318/vshow}{GSS Data Explorer page for \code{supduty1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 1176 \tab 200 \tab 130 \tab 11 \tab 1517 \cr
   Total \tab 1176 \tab 200 \tab 130 \tab 11 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supothrs}},
\code{\link{totsup}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
