% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_141.R
\name{supduty5}
\alias{supduty5}
\title{You check workers' progress, attendance}
\source{
General Social Survey https://gss.norc.org
}
\description{
supduty5
}
\details{
Question Here is a card listing 7 different kinds of oversight activities. Which ones do you do as part of your supervisory responsibilities?  Please mention all that apply:
E. Check workers' progress or attendance.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1322/vshow}{GSS Data Explorer page for \code{supduty5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 1176 \tab 253 \tab 77 \tab 11 \tab 1517 \cr
   Total \tab 1176 \tab 253 \tab 77 \tab 11 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty6}},
\code{\link{supduty7}}
}
\concept{Work Organizations}
\keyword{variable}
