% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_141.R
\name{supduty7}
\alias{supduty7}
\title{You take promotions or disciplinary actions}
\source{
General Social Survey https://gss.norc.org
}
\description{
supduty7
}
\details{
Question Here is a card listing 7 different kinds of oversight activities. Which ones do you do as part of your supervisory responsibilities?  Please mention all that apply:
G. Recommend promotions or disciplinary actions.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1324/vshow}{GSS Data Explorer page for \code{supduty7}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 1176 \tab 195 \tab 135 \tab 11 \tab 1517 \cr
   Total \tab 1176 \tab 195 \tab 135 \tab 11 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}}
}
\concept{Work Organizations}
\keyword{variable}
