% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_326.R
\name{supfam1}
\alias{supfam1}
\title{Aged should be able to support family}
\source{
General Social Survey https://gss.norc.org
}
\description{
supfam1
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important is it for them to be...
Be capable of supporting a family financially
A. By what age should this normally occur?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2905/vshow}{GSS Data Explorer page for \code{supfam1}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 35 \tab 37 \tab 40 \tab 55 \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1483 \tab 7 \tab 3 \tab 58 \tab 12 \tab 62 \tab 116 \tab 100 \tab 65 \tab 92 \tab 373 \tab 61 \tab 37 \tab 55 \tab 9 \tab 138 \tab 1 \tab 7 \tab 14 \tab 2 \tab 1 \tab 1 \tab 65 \tab 3 \tab 2765 \cr
   Total \tab 1483 \tab 7 \tab 3 \tab 58 \tab 12 \tab 62 \tab 116 \tab 100 \tab 65 \tab 92 \tab 373 \tab 61 \tab 37 \tab 55 \tab 9 \tab 138 \tab 1 \tab 7 \tab 14 \tab 2 \tab 1 \tab 1 \tab 65 \tab 3 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{boycott}},
\code{\link{conoffcl}},
\code{\link{getmar}},
\code{\link{getmar1}},
\code{\link{havchld}},
\code{\link{havchld1}},
\code{\link{protest}},
\code{\link{signpet}},
\code{\link{supfam}}
}
\concept{Adult Transitions}
\keyword{variable}
