% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{suphelp}
\alias{suphelp}
\title{Supervisor helpful to r in getting job done}
\source{
General Social Survey https://gss.norc.org
}
\description{
suphelp
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
L. My supervisor is helpful to me in getting the job done
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2807/vshow}{GSS Data Explorer page for \code{suphelp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 867 \tab 570 \tab 155 \tab 132 \tab 40 \tab 32 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 779 \tab 597 \tab 151 \tab 110 \tab 64 \tab 33 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 541 \tab 417 \tab 91 \tab 75 \tab 27 \tab 36 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 610 \tab 439 \tab 107 \tab 54 \tab 24 \tab 16 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 722 \tab 489 \tab 131 \tab 43 \tab 25 \tab 38 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 955 \tab 681 \tab 160 \tab 110 \tab 36 \tab 19 \tab 28 \tab 3544 \cr
   Total \tab 8345 \tab 4474 \tab 3193 \tab 795 \tab 524 \tab 216 \tab 174 \tab 28 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
