% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_654.R
\name{svyid1}
\alias{svyid1}
\title{Willingness to use video chat}
\source{
General Social Survey https://gss.norc.org
}
\description{
svyid1
}
\details{
Question How likely is it that you would be willing to use a video chat for a survey interview within the next few months?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7864/vshow}{GSS Data Explorer page for \code{svyid1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 424 \tab 1015 \tab 825 \tab 1152 \tab 27 \tab 66 \tab 25 \tab 10 \tab 3544 \cr
   2024 \tab - \tab 391 \tab 885 \tab 804 \tab 1127 \tab 27 \tab 56 \tab 11 \tab 8 \tab 3309 \cr
   Total \tab 0 \tab 815 \tab 1900 \tab 1629 \tab 2279 \tab 54 \tab 122 \tab 36 \tab 18 \tab 6853 \cr
}
}

\keyword{variable}
