% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{svyid2}
\alias{svyid2}
\title{Willingnes for in-person survey}
\source{
General Social Survey https://gss.norc.org
}
\description{
svyid2
}
\details{
Question How likely is it that you would be willing to have someone visit your home for a survey in-person interview within the next few months?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7865/vshow}{GSS Data Explorer page for \code{svyid2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 66 \tab 441 \tab 1002 \tab 734 \tab 1231 \tab 32 \tab 25 \tab 13 \tab 3544 \cr
   Total \tab 66 \tab 441 \tab 1002 \tab 734 \tab 1231 \tab 32 \tab 25 \tab 13 \tab 3544 \cr
}
}

\keyword{variable}
