% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{swayvote}
\alias{swayvote}
\title{Has r tried to influence votes of others}
\source{
General Social Survey https://gss.norc.org
}
\description{
swayvote
}
\details{
Question During elections, do you ever try to show people why they should vote for one of the parties or candidates? Do you do that often, sometimes, rarely or never?
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1028/vshow}{GSS Data Explorer page for \code{swayvote}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 175 \tab 398 \tab 247 \tab 986 \tab 4 \tab 9 \tab 1819 \cr
   Total \tab 0 \tab 175 \tab 398 \tab 247 \tab 986 \tab 4 \tab 9 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
