% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{switage2}
\alias{switage2}
\title{How old was r when second switch}
\source{
General Social Survey https://gss.norc.org
}
\description{
switage2
}
\details{
Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/DENOM16).
Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION MENTIONED IN RELIG16/DENOM16).
Yes................(ASK A)
No..............(GO TO MARELIG)
C1. How old were you when you switched to (FIRST/SECOND MENTION)?
(2) Age at Second Switch
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1055/vshow}{GSS Data Explorer page for \code{switage2}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 11 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 39 \tab 40 \tab 41 \tab 45 \tab 46 \tab 50 \tab 55 \tab 57 \tab 60 \tab 61 \tab 67 \tab 73 \tab don't know \tab no answer \tab Total \cr
   1988 \tab 1200 \tab 1 \tab 3 \tab 6 \tab 4 \tab 4 \tab 6 \tab 4 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 205 \tab 1481 \cr
   Total \tab 1200 \tab 1 \tab 3 \tab 6 \tab 4 \tab 4 \tab 6 \tab 4 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 205 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{maden}},
\code{\link{marelig}},
\code{\link{switage1}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}}
}
\concept{Religion}
\keyword{variable}
