% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{switch3}
\alias{switch3}
\title{What was r's third religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
switch3
}
\details{
Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/ DENOM16).
Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION menTIONED IN RELIG16/DENOM16).
Yes................(ASK A)
No..............(GO TO MARELIG)
A. Starting when you were 16 years old, what other religious preferences have you had?  Please mention your earliest preference first. Any others?
3. Third Mention
}
\section{Values}{

\itemize{
\item \verb{[110000]} amer baptist assoc
\item \verb{[111000]} amer bapt ch in us
\item \verb{[112000]} natl bapt con of am
\item \verb{[113000]} natl bapt con, usa
\item \verb{[114000]} southern bapt con
\item \verb{[115000]} other baptist church
\item \verb{[118000]} baptist, dk which
\item \verb{[120000]} afrcn meth epis ch
\item \verb{[121000]} afrcn meth epis zion
\item \verb{[122000]} united methodist ch
\item \verb{[123000]} other methodist ch
\item \verb{[128000]} methodist, dk which
\item \verb{[130000]} amer lutheran church
\item \verb{[131000]} lutheran ch in amer
\item \verb{[132000]} luth ch-missouri syn
\item \verb{[133000]} wisconsin ev luth sn
\item \verb{[134000]} other lutheran churc
\item \verb{[135000]} evangelical lutheran
\item \verb{[138000]} lutheran, dk which
\item \verb{[140000]} presbyterian ch n us
\item \verb{[141000]} united pres ch in us
\item \verb{[142000]} other presbyterian
\item \verb{[148000]} presbyterian, dk whi
\item \verb{[150000]} episcopal church
\item \verb{[160000]} other protestant
\item \verb{[170000]} no denomination
\item \verb{[2e+05]} catholic
\item \verb{[3e+05]} jewish
\item \verb{[4e+05]} none
\item \verb{[5e+05]} other religion
\item \verb{[999999]} missing
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1052/vshow}{GSS Data Explorer page for \code{switch3}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab southern bapt con \tab baptist, dk which \tab united methodist ch \tab other methodist ch \tab methodist, dk which \tab lutheran ch in amer \tab wisconsin ev luth sn \tab lutheran, dk which \tab united pres ch in us \tab presbyterian, dk whi \tab 160003 \tab 160035 \tab 160040 \tab 160068 \tab no denomination \tab catholic \tab jewish \tab other religion \tab no answer \tab Total \cr
   1988 \tab 1297 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 154 \tab 1481 \cr
   Total \tab 1297 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 154 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{maden}},
\code{\link{marelig}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}}
}
\concept{Religion}
\keyword{variable}
