% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{switched}
\alias{switched}
\title{Has r ever had a different religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
switched
}
\details{
Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN   RELIG16/ DENOM16).
Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION menTIONED IN RELIG16/ DENOM16).
Yes................(ASK A)
No..............(GO TO MARELIG)
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1049/vshow}{GSS Data Explorer page for \code{switched}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1988 \tab - \tab 525 \tab 945 \tab 11 \tab 1481 \cr
   Total \tab 0 \tab 525 \tab 945 \tab 11 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}}
}
\concept{Religion}
\keyword{variable}
