% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_243.R
\name{swithlth}
\alias{swithlth}
\title{Ever switched plan medical care provider}
\source{
General Social Survey https://gss.norc.org
}
\description{
swithlth
}
\details{
Question Have you ever switched your plan or the medical care provider within your plan due to dissatisfaction with a prior plan?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} cant switch
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2188/vshow}{GSS Data Explorer page for \code{swithlth}}.
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{mntlcare}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid3}},
\code{\link{psycpsyc}},
\code{\link{usedmntl}}
}
\concept{Mental Health}
\keyword{variable}
