% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{switnum}
\alias{switnum}
\title{How many religious preferences mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
switnum
}
\details{
Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/DENOM16).
Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION MENTIONED IN RELIG16/DENOM16).
Yes................(ASK A)
No..............(GO TO MARELIG)
B. RECORD NUMBER OF RELIGIONS MENTIONED
}
\section{Values}{

\itemize{
\item \verb{[4]} four or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1053/vshow}{GSS Data Explorer page for \code{switnum}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab four or more \tab no answer \tab Total \cr
   1988 \tab 945 \tab 255 \tab 97 \tab 24 \tab 6 \tab 154 \tab 1481 \cr
   Total \tab 945 \tab 255 \tab 97 \tab 24 \tab 6 \tab 154 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{maden}},
\code{\link{marelig}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switwhy1}},
\code{\link{switwhy2}}
}
\concept{Religion}
\keyword{variable}
