% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_419.R
\name{synmem}
\alias{synmem}
\title{R belongs to a synagogue, temple, minyan, or high holiday congregation}
\source{
General Social Survey https://gss.norc.org
}
\description{
synmem
}
\details{
Question (Do you/Does someone in your household) belong to a synagogue, temple, minyan, or high holiday congregation?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3829/vshow}{GSS Data Explorer page for \code{synmem}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 1968 \tab 24 \tab 31 \tab 2023 \cr
   2012 \tab 1928 \tab 15 \tab 31 \tab 1974 \cr
   Total \tab 3896 \tab 39 \tab 62 \tab 3997 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{bmitzvah}},
\code{\link{kd1relig}},
\code{\link{kd2relig}},
\code{\link{kd3relig}},
\code{\link{kd4relig}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}}
}
\concept{Jewish Identity}
\keyword{variable}
