% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{tablprce}
\alias{tablprce}
\title{The factory not lowering price acceptable?}
\source{
General Social Survey https://gss.norc.org
}
\description{
tablprce
}
\details{
Question Consider the following situation: A small factory produces tables and sells all that it can make at $200 each. because of changes in the price of materials, the cost of making each table has recently decreased by $20. The factory does not lower its price for the tables. Is this acceptable or unfair?
}
\section{Values}{

\itemize{
\item \code{1} acceptable
\item \code{2} unfair
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1865/vshow}{GSS Data Explorer page for \code{tablprce}}.
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
