% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{tagsales}
\alias{tagsales}
\title{How frequently do you go to garage sales etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
tagsales
}
\details{
Question How frequently do you go to garage sales, tag sales, or yard sales -- that is, one-time sales in which homeowners sell items that they no longer need from out of their homes, yards, or garages?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} rarely
\item \verb{[3]} occasionally 1-3 times a yr
\item \verb{[4]} regularly
\item \verb{[5]} almost every weekend
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1874/vshow}{GSS Data Explorer page for \code{tagsales}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab rarely \tab occasionally 1-3 times a yr \tab regularly \tab almost every weekend \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 457 \tab 397 \tab 382 \tab 146 \tab 56 \tab 6 \tab 2904 \cr
   Total \tab 1460 \tab 457 \tab 397 \tab 382 \tab 146 \tab 56 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
