% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{taketrck}
\alias{taketrck}
\title{Punishment for stealing and selling truck parts}
\source{
General Social Survey https://gss.norc.org
}
\description{
taketrck
}
\details{
Question Now I'm going to ask you about several offenses. For each case, indicate (Number 1-6) what you would consider the appropriate punishment for each case.
E. A government employee was convicted of stealing truck parts and tires from a military depot and selling them to civilians.
}
\section{Values}{

\itemize{
\item \verb{[1]} reprimand
\item \verb{[2]} dismiss military
\item \verb{[3]} one year prison
\item \verb{[4]} 5-9 years prison
\item \verb{[5]} 10-20 yrs prison
\item \verb{[6]} life in prison
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5231/vshow}{GSS Data Explorer page for \code{taketrck}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab reprimand \tab dismiss military \tab one year prison \tab 5-9 years prison \tab 10-20 yrs prison \tab life in prison \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 96 \tab 447 \tab 366 \tab 303 \tab 87 \tab 20 \tab 50 \tab 18 \tab 2832 \cr
   Total \tab 1445 \tab 96 \tab 447 \tab 366 \tab 303 \tab 87 \tab 20 \tab 50 \tab 18 \tab 2832 \cr
}
}

\seealso{
Other National Security: 
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{punarms}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}}
}
\concept{National Security}
\keyword{variable}
