% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{talkpol}
\alias{talkpol}
\title{How often does r discuss politics?}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkpol
}
\details{
Question Thinking of the (people/person) we''ve just been discussing, would you say you discuss social and political issues with (him/her/them) almost all the time, most of the time, occasionally, or almost never?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} most of the time
\item \code{3} occasionally
\item \code{4} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/961/vshow}{GSS Data Explorer page for \code{talkpol}}.
}

\seealso{
Other Social Networks: 
\code{\link{frndknow}},
\code{\link{talkpol3}}
}
\concept{Social Networks}
\keyword{variable}
