% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_104.R
\name{talkpol1}
\alias{talkpol1}
\title{R discusses politics with person #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkpol1
}
\details{
Question About how often do you talk to (NAME) about political matters?
\enumerate{
\item Political Talks with First Person
}
}
\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} at least weekly
\item \code{3} at least monthly
\item \code{4} at least yearly
\item \code{5} less than yearly
\item \code{6} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/958/vshow}{GSS Data Explorer page for \code{talkpol1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab almost daily \tab at least weekly \tab at least monthly \tab at least yearly \tab less than yearly \tab never \tab don't know \tab no answer \tab Total \cr
   1987 \tab 98 \tab 245 \tab 365 \tab 373 \tab 262 \tab 160 \tab 284 \tab 5 \tab 27 \tab 1819 \cr
   Total \tab 98 \tab 245 \tab 365 \tab 373 \tab 262 \tab 160 \tab 284 \tab 5 \tab 27 \tab 1819 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{talkpol2}}
}
\concept{Social Networks}
\keyword{variable}
