% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_104.R
\name{talkpol2}
\alias{talkpol2}
\title{R discusses politics with person #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkpol2
}
\details{
Question About how often do you talk to (NAME) about political matters?
2. Political Talks with Second Person
}
\section{Values}{

\itemize{
\item \verb{[1]} almost daily
\item \verb{[2]} at least weekly
\item \verb{[3]} at least monthly
\item \verb{[4]} at least yearly
\item \verb{[5]} less than yearly
\item \verb{[6]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/959/vshow}{GSS Data Explorer page for \code{talkpol2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab almost daily \tab at least weekly \tab at least monthly \tab at least yearly \tab less than yearly \tab never \tab don't know \tab no answer \tab Total \cr
   1987 \tab 385 \tab 103 \tab 245 \tab 363 \tab 269 \tab 153 \tab 272 \tab 3 \tab 26 \tab 1819 \cr
   Total \tab 385 \tab 103 \tab 245 \tab 363 \tab 269 \tab 153 \tab 272 \tab 3 \tab 26 \tab 1819 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{talkpol1}}
}
\concept{Social Networks}
\keyword{variable}
