% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{talkpol3}
\alias{talkpol3}
\title{R discusses politics with person #3}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkpol3
}
\details{
Question About how often do you talk to (NAME) about political matters?
3. Political Talks with Third Person
}
\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} at least weekly
\item \code{3} at least monthly
\item \code{4} at least yearly
\item \code{5} less than yearly
\item \code{6} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/960/vshow}{GSS Data Explorer page for \code{talkpol3}}.
}

\seealso{
Other Social Networks: 
\code{\link{frndknow}},
\code{\link{talkpol}}
}
\concept{Social Networks}
\keyword{variable}
