% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{talkspvs}
\alias{talkspvs}
\title{Comfortable talking with supervisor about personal}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkspvs
}
\details{
Question How comfortable are you talking with your supervisor about conflicts between work and personal or family life
}
\section{Values}{

\itemize{
\item \code{1} not at all comfortable
\item \code{2} a little
\item \code{3} somewhat
\item \code{4} very
\item \code{5} extremely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5505/vshow}{GSS Data Explorer page for \code{talkspvs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all comfortable \tab a little \tab somewhat \tab very \tab extremely \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1505 \tab 139 \tab 119 \tab 262 \tab 337 \tab 166 \tab 5 \tab 5 \tab 2538 \cr
   Total \tab 1505 \tab 139 \tab 119 \tab 262 \tab 337 \tab 166 \tab 5 \tab 5 \tab 2538 \cr
}
}

\seealso{
Other Work and Relationships: 
\code{\link{effctsup}},
\code{\link{wkcontct}}
}
\concept{Work and Relationships}
\keyword{variable}
