% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{talksup}
\alias{talksup}
\title{How likely to speak supervisor if coworker not w}
\source{
General Social Survey https://gss.norc.org
}
\description{
talksup
}
\details{
Question If you were to see a fellow employee not working as hard or well as he or she should, how likely would you be to:
B. Speak to your supervisor or manager
}
\section{Values}{

\itemize{
\item \code{0} i do not have a supervisor or manager
\item \code{1} not at all likely
\item \code{2} not very likely
\item \code{3} somewhat likely
\item \code{4} very likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2866/vshow}{GSS Data Explorer page for \code{talksup}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{cowrkhrd}},
\code{\link{donothng}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{talkemp}}

Other Job: 
\code{\link{cowrkhrd}},
\code{\link{donothng}},
\code{\link{talkemp}}
}
\concept{Employee Compensation}
\concept{Job}
\keyword{variable}
