% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_102.R
\name{talkto2}
\alias{talkto2}
\title{How often does r talk to person #2?}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkto2
}
\details{
Question Thinking about how often you usually talk to (NAME), on the average, Do you talk to (him/her) almost every day, at least Once a week, at least once a month, or less than once a month? ASK FOR EACH NAME.
2. Talks to Second Person
}
\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} once a week
\item \code{3} once a month
\item \code{4} lt once a month
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/931/vshow}{GSS Data Explorer page for \code{talkto2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab almost daily \tab once a week \tab once a month \tab lt once a month \tab no answer \tab Total \cr
   1985 \tab 364 \tab 583 \tab 389 \tab 144 \tab 48 \tab 6 \tab 1534 \cr
   2004 \tab 1977 \tab 377 \tab 304 \tab 88 \tab 19 \tab 47 \tab 2812 \cr
   Total \tab 2341 \tab 960 \tab 693 \tab 232 \tab 67 \tab 53 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}}
}
\concept{Social Networks}
\keyword{variable}
