% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_102.R
\name{talkto3}
\alias{talkto3}
\title{How often does r talk to person #3?}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkto3
}
\details{
Question Thinking about how often you usually talk to (NAME), on the average, Do you talk to (him/her) almost every day, at least Once a week, at least once a month, or less than once a month? ASK FOR EACH NAME.
3. Talks to Third Person
}
\section{Values}{

\itemize{
\item \verb{[1]} almost daily
\item \verb{[2]} once a week
\item \verb{[3]} once a month
\item \verb{[4]} lt once a month
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/932/vshow}{GSS Data Explorer page for \code{talkto3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab almost daily \tab once a week \tab once a month \tab lt once a month \tab no answer \tab Total \cr
   1985 \tab 599 \tab 377 \tab 354 \tab 144 \tab 53 \tab 7 \tab 1534 \cr
   2004 \tab 2240 \tab 220 \tab 211 \tab 75 \tab 19 \tab 47 \tab 2812 \cr
   Total \tab 2839 \tab 597 \tab 565 \tab 219 \tab 72 \tab 54 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto4}},
\code{\link{talkto5}}
}
\concept{Social Networks}
\keyword{variable}
