% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{taphme}
\alias{taphme}
\title{Govt wire-tap home phone line}
\source{
General Social Survey https://gss.norc.org
}
\description{
taphme
}
\details{
Question Protecting secrets is a continuing concern of the government. I am going to ask you about various security measures that the government might apply to individuals with a SECRET or TOP SECRET clearance. Do you (strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree) that people with SECRET or TOP SECRET clearance should be subject to:
F. Wire-tapping of telephone calls at home
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5254/vshow}{GSS Data Explorer page for \code{taphme}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 106 \tab 173 \tab 188 \tab 534 \tab 366 \tab 34 \tab 18 \tab 2817 \cr
   Total \tab 1398 \tab 106 \tab 173 \tab 188 \tab 534 \tab 366 \tab 34 \tab 18 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{camwrk}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{nuclrwar}},
\code{\link{srchwrk}},
\code{\link{tapwrk}},
\code{\link{usspy}},
\code{\link{usterror}}
}
\concept{National Security}
\keyword{variable}
