% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{taxmid}
\alias{taxmid}
\title{Taxes on middle income people too high}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxmid
}
\details{
Question Generally, how would you describe taxes in America today...
We mean all taxes together, including social security, income tax, sales tax, and all the rest.
B. Next, for those with middle incomes, are taxes . . .
}
\section{Values}{

\itemize{
\item \code{1} much too high
\item \code{2} too high
\item \code{3} about right
\item \code{4} too low
\item \code{5} much too low
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4303/vshow}{GSS Data Explorer page for \code{taxmid}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}

Other Taxes: 
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\concept{Taxes}
\keyword{variable}
