% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{taxpaid}
\alias{taxpaid}
\title{Do authorities ensure taxes are paid}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxpaid
}
\details{
Question In general, how often do you think that the tax authorities in America do the following?
Make sure people pay their taxes?  Would you say...
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6311/vshow}{GSS Data Explorer page for \code{taxpaid}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
