% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{taxserve}
\alias{taxserve}
\title{Favor conscription given taxes?}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxserve
}
\details{
Question C. IF STRONGLY FAVOR OR PROBABLY FAVOR EITHER A OR B: and suppose that the costs of such a program made it necessary to increase your taxes by a small amount--for example, 5 percent. Would you strongly favor it, probably favor it, probably oppose it, or strongly oppose it?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} probably favor
\item \code{3} probably oppose
\item \code{4} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/780/vshow}{GSS Data Explorer page for \code{taxserve}}.
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}}

Other Military: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}}
}
\concept{Core}
\concept{Military}
\concept{Split Ballots}
\keyword{variable}
