% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{taxspend}
\alias{taxspend}
\title{Choice between reducing taxes or spending more}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxspend
}
\details{
Question If the government had a choice between reducing taxes or spending more on social programs like health care, social security, and unemployment benefits, which do you think it should do? (We mean all taxes together, including social security, income tax, sales tax, and all the rest.) PLEASE CHECK ONE BOX ONLY.
}
\section{Values}{

\itemize{
\item \code{1} reduce taxes
\item \code{2} spend more on social programs
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4103/vshow}{GSS Data Explorer page for \code{taxspend}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}}

Other Federal Government: 
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}}
}
\concept{Federal Government}
\concept{ISSP Role of Gov't}
\concept{Spending}
\concept{Taxes}
\keyword{variable}
