% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{teacheth}
\alias{teacheth}
\title{Ethnicity of college faculty reflect students}
\source{
General Social Survey https://gss.norc.org
}
\description{
teacheth
}
\details{
Question Some people say that it is only fair that in a society like ours, people holding public office should have roughly the same racial or ethnic background as the population as a whole. So, for example, if 30pct of the people were Hispanic, then 30pct of the public officials should be Hispanic as well. Others say that in choosing people for public office they should be considered purely on the basis of ability without regard to their ethnic or racial background. Let me ask you about these ideas in several different areas.
B. What about teachers in public schools or universities?  Should they have roughly the same racial or ethnic background as their students, or should they be considered purely on the basis of their ability without regard to their racial or ethnic background?
}
\section{Values}{

\itemize{
\item \code{1} consider ethnic background
\item \code{2} ability, not background
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1628/vshow}{GSS Data Explorer page for \code{teacheth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab consider ethnic background \tab ability, not background \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 115 \tab 1310 \tab 37 \tab 12 \tab 2992 \cr
   Total \tab 1518 \tab 115 \tab 1310 \tab 37 \tab 12 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{ethhist}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
