% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{teamsafe}
\alias{teamsafe}
\title{Mgt and employees work together re safety}
\source{
General Social Survey https://gss.norc.org
}
\description{
teamsafe
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
M. Where I work, employees and management work together to ensure the safest possible working conditions
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2786/vshow}{GSS Data Explorer page for \code{teamsafe}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 652 \tab 900 \tab 168 \tab 37 \tab 23 \tab 16 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 631 \tab 882 \tab 128 \tab 26 \tab 47 \tab 20 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 411 \tab 630 \tab 85 \tab 18 \tab 14 \tab 29 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 477 \tab 658 \tab 81 \tab 13 \tab 12 \tab 9 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 464 \tab 812 \tab 107 \tab 16 \tab 16 \tab 33 \tab - \tab 2348 \cr
   2022 \tab 1547 \tab 732 \tab 990 \tab 171 \tab 40 \tab 24 \tab 14 \tab 26 \tab 3544 \cr
   Total \tab 8337 \tab 3367 \tab 4872 \tab 740 \tab 150 \tab 136 \tab 121 \tab 26 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
