% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{techesy}
\alias{techesy}
\title{Does technology make our lives easier?}
\source{
General Social Survey https://gss.norc.org
}
\description{
techesy
}
\details{
Question Here are some statements about technology. Technology includes all kinds of machinery, equipment, digital devices, and software. How much do you agree or disagree with each of these statements?… Technology makes our lives easier.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8061/vshow}{GSS Data Explorer page for \code{techesy}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 467 \tab 798 \tab 233 \tab 58 \tab 19 \tab 53 \tab 1668 \tab 3 \tab 10 \tab 3309 \cr
   Total \tab 0 \tab 467 \tab 798 \tab 233 \tab 58 \tab 19 \tab 53 \tab 1668 \tab 3 \tab 10 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{intcnfrm}},
\code{\link{inthrss}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intrust}},
\code{\link{intscam}},
\code{\link{intviews}},
\code{\link{nextgen1}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
