% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{techjobs}
\alias{techjobs}
\title{New techs increase number of jobs}
\source{
General Social Survey https://gss.norc.org
}
\description{
techjobs
}
\details{
Question New kinds of technology are being introduced more and more in the United States: computers, robots, and so on. Do you think these new technologies will over the next few years...
}
\section{Values}{

\itemize{
\item \code{1} greatly increase
\item \code{2} slight increase
\item \code{3} no difference to
\item \code{4} slightly reduce
\item \code{5} greatly reduce
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4575/vshow}{GSS Data Explorer page for \code{techjobs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab greatly increase \tab slight increase \tab no difference to \tab slightly reduce \tab greatly reduce \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1609 \tab 183 \tab 238 \tab 129 \tab 329 \tab 232 \tab 87 \tab 25 \tab 2832 \cr
   Total \tab 1609 \tab 183 \tab 238 \tab 129 \tab 329 \tab 232 \tab 87 \tab 25 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techwork}},
\code{\link{wantjob1}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
