% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{techwork}
\alias{techwork}
\title{New techs make work more interesting}
\source{
General Social Survey https://gss.norc.org
}
\description{
techwork
}
\details{
Question Do you think that the introduction of new technologies in the United States over the next few years will make work…
}
\section{Values}{

\itemize{
\item \verb{[1]} much more
\item \verb{[2]} little more
\item \verb{[3]} neither mre/less
\item \verb{[4]} a little less
\item \verb{[5]} much less
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4576/vshow}{GSS Data Explorer page for \code{techwork}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much more \tab little more \tab neither mre/less \tab a little less \tab much less \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1609 \tab 349 \tab 411 \tab 224 \tab 87 \tab 45 \tab 84 \tab 23 \tab 2832 \cr
   Total \tab 1609 \tab 349 \tab 411 \tab 224 \tab 87 \tab 45 \tab 84 \tab 23 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{wantjob1}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
