% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_15.R
\name{teens}
\alias{teens}
\title{Household members 13 thru 17 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
teens
}
\details{
Question Household Size and Composition (see Appendix D:  Recodes, for information about these variables). C. NUMBER OF MEMBERS 13 TO 17 YEARS OLD:
As of 2024, TEENS has been recoded to only account for 0 to 2+ teenagers in the household. This has been retroactively applied to all previous years of TEENS. The original TEENS containing data from 1972 to 2022 has been renamed TEENS_7222.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0
\item \verb{[1]} 1
\item \verb{[2]} 2 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/100/vshow}{GSS Data Explorer page for \code{teens}}.

Counts by year:\tabular{llllllll}{
   year \tab 0 \tab 1 \tab 2 or more \tab no answer \tab skipped on web \tab iap \tab Total \cr
   1972 \tab 1182 \tab 256 \tab 175 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1156 \tab 220 \tab 128 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1112 \tab 215 \tab 155 \tab 2 \tab - \tab - \tab 1484 \cr
   1975 \tab 1152 \tab 207 \tab 131 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1189 \tab 182 \tab 126 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab 1190 \tab 190 \tab 142 \tab 8 \tab - \tab - \tab 1530 \cr
   1978 \tab 1204 \tab 200 \tab 120 \tab 8 \tab - \tab - \tab 1532 \cr
   1980 \tab 1208 \tab 164 \tab 86 \tab 10 \tab - \tab - \tab 1468 \cr
   1982 \tab 1535 \tab 215 \tab 107 \tab 3 \tab - \tab - \tab 1860 \cr
   1983 \tab 1301 \tab 210 \tab 84 \tab 4 \tab - \tab - \tab 1599 \cr
   1984 \tab 1220 \tab 169 \tab 82 \tab 2 \tab - \tab - \tab 1473 \cr
   1985 \tab 1286 \tab 170 \tab 75 \tab 3 \tab - \tab - \tab 1534 \cr
   1986 \tab 1209 \tab 186 \tab 70 \tab 5 \tab - \tab - \tab 1470 \cr
   1987 \tab 1489 \tab 227 \tab 101 \tab 2 \tab - \tab - \tab 1819 \cr
   1988 \tab 1252 \tab 168 \tab 55 \tab 6 \tab - \tab - \tab 1481 \cr
   1989 \tab 1254 \tab 210 \tab 58 \tab 15 \tab - \tab - \tab 1537 \cr
   1990 \tab 1191 \tab 128 \tab 46 \tab 7 \tab - \tab - \tab 1372 \cr
   1991 \tab 1315 \tab 146 \tab 52 \tab 4 \tab - \tab - \tab 1517 \cr
   1993 \tab 1377 \tab 173 \tab 56 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2567 \tab 326 \tab 98 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 2492 \tab 306 \tab 106 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2414 \tab 321 \tab 96 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 2360 \tab 316 \tab 98 \tab 43 \tab - \tab - \tab 2817 \cr
   2002 \tab 2437 \tab 230 \tab 68 \tab 30 \tab - \tab - \tab 2765 \cr
   2004 \tab 2427 \tab 256 \tab 113 \tab 16 \tab - \tab - \tab 2812 \cr
   2006 \tab 3835 \tab 470 \tab 173 \tab 32 \tab - \tab - \tab 4510 \cr
   2008 \tab 1711 \tab 207 \tab 75 \tab 30 \tab - \tab - \tab 2023 \cr
   2010 \tab 1794 \tab 170 \tab 75 \tab 5 \tab - \tab - \tab 2044 \cr
   2012 \tab 1713 \tab 185 \tab 56 \tab 20 \tab - \tab - \tab 1974 \cr
   2014 \tab 2250 \tab 197 \tab 78 \tab 13 \tab - \tab - \tab 2538 \cr
   2016 \tab 2520 \tab 237 \tab 93 \tab 17 \tab - \tab - \tab 2867 \cr
   2018 \tab 2074 \tab 186 \tab 64 \tab 24 \tab - \tab - \tab 2348 \cr
   2021 \tab 3622 \tab 293 \tab 114 \tab - \tab 3 \tab - \tab 4032 \cr
   2022 \tab 1601 \tab 108 \tab 59 \tab 8 \tab 4 \tab 1764 \tab 3544 \cr
   2024 \tab 1493 \tab 124 \tab 59 \tab 1 \tab 2 \tab 1630 \tab 3309 \cr
   Total \tab 61132 \tab 7568 \tab 3274 \tab 322 \tab 9 \tab 3394 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{granborn}},
\code{\link{hompop}},
\code{\link{income}},
\code{\link{preteen}},
\code{\link{rincome}},
\code{\link{unrelat}}

Other Household: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{hompop}},
\code{\link{preteen}},
\code{\link{unrelat}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
