% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{teens_exp}
\alias{teens_exp}
\title{Kids 13-17 in hh(experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
teens_exp
}
\details{
Question How many individuals are there in this household between the ages of 13 and 17?
As of 2024, TEENS_EXP has been recoded to only account for 0 to 2+ teenagers in the household. This has been retroactively applied to all previous years of TEENS_EXP. The original TEENS_EXP containing data from in 2022 has been renamed TEENS_EXP_22.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0
\item \verb{[1]} 1
\item \verb{[2]} 2 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8011/vshow}{GSS Data Explorer page for \code{teens_exp}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 or more \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 1524 \tab 152 \tab 52 \tab 1788 \tab 10 \tab 18 \tab 3544 \cr
   2024 \tab - \tab 1388 \tab 162 \tab 51 \tab 1684 \tab 3 \tab 21 \tab 3309 \cr
   Total \tab 0 \tab 2912 \tab 314 \tab 103 \tab 3472 \tab 13 \tab 39 \tab 6853 \cr
}
}

\seealso{
Other Core: 
\code{\link{babies_exp}},
\code{\link{preteen_exp}}

Other Household: 
\code{\link{babies_exp}},
\code{\link{preteen_exp}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
