% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{tempgen1}
\alias{tempgen1}
\title{Greenhouse effect danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
tempgen1
}
\details{
Question H. In general, do you think that a rise in the world's temperature caused by climate change is . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4810/vshow}{GSS Data Explorer page for \code{tempgen1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 286 \tab 364 \tab 416 \tab 181 \tab 78 \tab 100 \tab 5 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 528 \tab 501 \tab 493 \tab 166 \tab 46 \tab 98 \tab 187 \tab 10 \tab 4032 \cr
   Total \tab 2617 \tab 814 \tab 865 \tab 909 \tab 347 \tab 124 \tab 198 \tab 192 \tab 10 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{peopgrn}},
\code{\link{redcehme}}
}
\concept{ISSP Environment}
\keyword{variable}
