% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{termemot}
\alias{termemot}
\title{Worry about emotional burden on my family}
\source{
General Social Survey https://gss.norc.org
}
\description{
termemot
}
\details{
Question Now, I would like to talk about concerns you may have when considering what may happen at the end of your life. please tell me how much you agree or disagree with each of these statements:
C. I worry about the emotional burden that my family might face making decisions for me at the end of life.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2248/vshow}{GSS Data Explorer page for \code{termemot}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 283 \tab 566 \tab 118 \tab 311 \tab 79 \tab 22 \tab 8 \tab 2832 \cr
   Total \tab 1445 \tab 283 \tab 566 \tab 118 \tab 311 \tab 79 \tab 22 \tab 8 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
