% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{terminal}
\alias{terminal}
\title{Have you or close friends relative faced with terminal illness}
\source{
General Social Survey https://gss.norc.org
}
\description{
terminal
}
\details{
Question Now in this section, I will ask you some questions about the end of life. Have you or any close friend or family members been faced with a terminal or life-threatening illness?
}
\section{Values}{

\itemize{
\item \code{1} yes myself
\item \code{2} yes friend
\item \code{3} yes immed reltve
\item \code{4} yes othr relatve
\item \code{5} no
\item \code{6} self and one
\item \code{7} two or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2245/vshow}{GSS Data Explorer page for \code{terminal}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab yes myself \tab yes friend \tab yes immed reltve \tab yes othr relatve \tab no \tab self and one \tab two or more \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 71 \tab 100 \tab 324 \tab 370 \tab 459 \tab 7 \tab 26 \tab 14 \tab 16 \tab 2832 \cr
   Total \tab 1445 \tab 71 \tab 100 \tab 324 \tab 370 \tab 459 \tab 7 \tab 26 \tab 14 \tab 16 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
