% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{termpay}
\alias{termpay}
\title{Worry if no money get second class care}
\source{
General Social Survey https://gss.norc.org
}
\description{
termpay
}
\details{
Question Now, I would like to talk about concerns you may have when considering what may happen at the end of your life. please tell me how much you agree or disagree with each of these statements:
E. I worry that if I run out of money or health insurance I will get second class health care.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2250/vshow}{GSS Data Explorer page for \code{termpay}}.
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
