% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{termrel}
\alias{termrel}
\title{My religious community very helpful}
\source{
General Social Survey https://gss.norc.org
}
\description{
termrel
}
\details{
Question Now, I would like to talk about concerns you may have when considering what may happen at the end of your life. please tell me how much you agree or disagree with each of these statements:
B. My religious community would be very helpful if I were terminally ill.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2247/vshow}{GSS Data Explorer page for \code{termrel}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 236 \tab 510 \tab 199 \tab 228 \tab 134 \tab 66 \tab 14 \tab 2832 \cr
   Total \tab 1445 \tab 236 \tab 510 \tab 199 \tab 228 \tab 134 \tab 66 \tab 14 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}}
}
\concept{Mental Health}
\keyword{variable}
