% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{termvig}
\alias{termvig}
\title{One of the vignettes}
\source{
General Social Survey https://gss.norc.org
}
\description{
termvig
}
\details{
Question Now I'm going to describe a patient to you and then ask you some questions about that patient.
}
\section{Values}{

\itemize{
\item \verb{[1]} vignette a
\item \verb{[2]} vignette b
\item \verb{[3]} vignette c
\item \verb{[4]} vignette d
\item \verb{[5]} vignette e
\item \verb{[6]} vignette f
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2254/vshow}{GSS Data Explorer page for \code{termvig}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab vignette a \tab vignette b \tab vignette c \tab vignette d \tab vignette e \tab vignette f \tab Total \cr
   1998 \tab 1445 \tab 223 \tab 224 \tab 231 \tab 246 \tab 231 \tab 232 \tab 2832 \cr
   Total \tab 1445 \tab 223 \tab 224 \tab 231 \tab 246 \tab 231 \tab 232 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{painmed}},
\code{\link{rightref}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\keyword{variable}
