% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_572.R
\name{testdrug}
\alias{testdrug}
\title{People with secret should be subject to drug tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
testdrug
}
\details{
Question Protecting secrets is a continuing concern of the government.
I am going to ask you about various security measures that the government might apply to individuals with a SECRET of TOP SECRET clearance. Do you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree that people with SECRET or TOP SECRET clearance should be subject to...
B. Random drug tests
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5244/vshow}{GSS Data Explorer page for \code{testdrug}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 618 \tab 624 \tab 49 \tab 46 \tab 10 \tab 25 \tab 15 \tab 2832 \cr
   2000 \tab 1398 \tab 687 \tab 544 \tab 92 \tab 45 \tab 12 \tab 22 \tab 17 \tab 2817 \cr
   Total \tab 2843 \tab 1305 \tab 1168 \tab 141 \tab 91 \tab 22 \tab 47 \tab 32 \tab 5649 \cr
}
}

\seealso{
Other National Security: 
\code{\link{bugging}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{compfin}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{lietest}}
}
\concept{National Security}
\keyword{variable}
