% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{theism}
\alias{theism}
\title{God concerned with human beings personally}
\source{
General Social Survey https://gss.norc.org
}
\description{
theism
}
\details{
Question Do you agree or disagree with the following. . .
A. There is a God who concerns Himself with every human being personally.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4657/vshow}{GSS Data Explorer page for \code{theism}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 585 \tab 333 \tab 199 \tab 72 \tab 55 \tab 65 \tab 50 \tab 1517 \cr
   1998 \tab 1548 \tab 493 \tab 370 \tab 142 \tab 107 \tab 67 \tab 50 \tab 55 \tab 2832 \cr
   2008 \tab 658 \tab 578 \tab 368 \tab 185 \tab 136 \tab 76 \tab 16 \tab 6 \tab 2023 \cr
   2018 \tab 1173 \tab 481 \tab 298 \tab 175 \tab 99 \tab 100 \tab 13 \tab 9 \tab 2348 \cr
   Total \tab 3537 \tab 2137 \tab 1369 \tab 701 \tab 414 \tab 298 \tab 144 \tab 120 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}}
}
\concept{ISSP Religion}
\keyword{variable}
