% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{thisjob1}
\alias{thisjob1}
\title{Heard job from a notices posted my workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
thisjob1
}
\details{
Question Which of the answers on this card describe how you heard about your present job with your employer (IF SELF-EMPLOYED:  current organization)? Please tell me all of the ways that you found out about this job.
A. A notice was posted or circulated at my work place.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1292/vshow}{GSS Data Explorer page for \code{thisjob1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 1243 \tab 42 \tab 200 \tab 32 \tab 1517 \cr
   Total \tab 1243 \tab 42 \tab 200 \tab 32 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobyears}},
\code{\link{promotng}},
\code{\link{samejob}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}
}
\concept{Work Organizations}
\keyword{variable}
