% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{thisjob2}
\alias{thisjob2}
\title{Heard job through my union}
\source{
General Social Survey https://gss.norc.org
}
\description{
thisjob2
}
\details{
Question Which of the answers on this card describe how you heard about your present job with your employer (IF SELF-EMPLOYED:  current organization)? Please tell me all of the ways that you found out about this job.
B. Through my union.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1293/vshow}{GSS Data Explorer page for \code{thisjob2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 1243 \tab 5 \tab 231 \tab 38 \tab 1517 \cr
   Total \tab 1243 \tab 5 \tab 231 \tab 38 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobyears}},
\code{\link{promotng}},
\code{\link{samejob}},
\code{\link{thisjob1}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}
}
\concept{Work Organizations}
\keyword{variable}
