% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{thisjob7}
\alias{thisjob7}
\title{Heard job others (specify)}
\source{
General Social Survey https://gss.norc.org
}
\description{
thisjob7
}
\details{
Question Which of the answers on this card describe how you heard about your present job with your employer (IF SELF-EMPLOYED:  current organization)? Please tell me all of the ways that you found out about this job.
G. Other (SPECIFY).
}
\section{Values}{

\itemize{
\item \verb{[1]} relative
\item \verb{[2]} friend
\item \verb{[3]} personnel
\item \verb{[7]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1298/vshow}{GSS Data Explorer page for \code{thisjob7}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab relative \tab friend \tab personnel \tab other \tab no answer \tab Total \cr
   1991 \tab 1243 \tab 5 \tab 191 \tab 1 \tab 24 \tab 53 \tab 1517 \cr
   Total \tab 1243 \tab 5 \tab 191 \tab 1 \tab 24 \tab 53 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobyears}},
\code{\link{promotng}},
\code{\link{samejob}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}}
}
\concept{Work Organizations}
\keyword{variable}
