% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{thnkangr}
\alias{thnkangr}
\title{How often thought about this situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
thnkangr
}
\details{
Question Even after they go away, feelings sometimes come back when we think about things. How often have you thought about this situation since it happened: very often, every once in a while or just once or twice, or never (until asked  about it just now)?
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} every once in a while
\item \verb{[3]} just once or twice
\item \verb{[4]} never-not once until now
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1766/vshow}{GSS Data Explorer page for \code{thnkangr}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab every once in a while \tab just once or twice \tab never-not once until now \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1936 \tab 140 \tab 292 \tab 292 \tab 219 \tab 25 \tab - \tab 2904 \cr
   Total \tab 1936 \tab 140 \tab 292 \tab 292 \tab 219 \tab 25 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{angrlast}},
\code{\link{angry}},
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{anxious}},
\code{\link{ashamed}},
\code{\link{atease}},
\code{\link{badbrks}},
\code{\link{beplesnt}},
\code{\link{calm}},
\code{\link{chnang1}},
\code{\link{chnang10}},
\code{\link{chnang11}},
\code{\link{chnang12}},
\code{\link{chnang13}},
\code{\link{chnang14}},
\code{\link{chnang15}},
\code{\link{chnang16}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{chnang6}},
\code{\link{chnang7}},
\code{\link{chnang8}},
\code{\link{chnang9}},
\code{\link{contentd}},
\code{\link{embarrss}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{hapfeel}},
\code{\link{howangry}},
\code{\link{likeoth}},
\code{\link{litcntrl}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{madat1}},
\code{\link{madat10}},
\code{\link{madat11}},
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat19}},
\code{\link{madat2}},
\code{\link{madat20}},
\code{\link{madat21}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat5}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}},
\code{\link{mostluck}},
\code{\link{noemote}},
\code{\link{noplan}},
\code{\link{notam}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{othresp}},
\code{\link{outraged}},
\code{\link{ovrjoyed}},
\code{\link{proud}},
\code{\link{reactok}},
\code{\link{restless}},
\code{\link{sad}},
\code{\link{selfresp}},
\code{\link{shakeblu}},
\code{\link{showangr}},
\code{\link{showfeel}},
\code{\link{usangry}},
\code{\link{usenthus}},
\code{\link{usfrustr}},
\code{\link{ushopefl}},
\code{\link{ussat}},
\code{\link{usupset}},
\code{\link{usworry}},
\code{\link{whnangry}},
\code{\link{worried}}

Other Feelings: 
\code{\link{accptoth}},
\code{\link{afailure}},
\code{\link{angrlast}},
\code{\link{angry}},
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{anxious}},
\code{\link{ashamed}},
\code{\link{atease}},
\code{\link{badbrks}},
\code{\link{beplesnt}},
\code{\link{big5a1}},
\code{\link{big5a2}},
\code{\link{big5b1}},
\code{\link{big5b2}},
\code{\link{big5c1}},
\code{\link{big5c2}},
\code{\link{big5d1}},
\code{\link{big5d2}},
\code{\link{big5e1}},
\code{\link{big5e2}},
\code{\link{calm}},
\code{\link{chnang1}},
\code{\link{chnang10}},
\code{\link{chnang11}},
\code{\link{chnang12}},
\code{\link{chnang13}},
\code{\link{chnang14}},
\code{\link{chnang15}},
\code{\link{chnang16}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{chnang6}},
\code{\link{chnang7}},
\code{\link{chnang8}},
\code{\link{chnang9}},
\code{\link{contentd}},
\code{\link{effort}},
\code{\link{embarrss}},
\code{\link{empathy1}},
\code{\link{empathy2}},
\code{\link{empathy3}},
\code{\link{empathy4}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{fidgety}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{hapfeel}},
\code{\link{hopeless}},
\code{\link{howangry}},
\code{\link{likeoth}},
\code{\link{litcntrl}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{madat1}},
\code{\link{madat10}},
\code{\link{madat11}},
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat19}},
\code{\link{madat2}},
\code{\link{madat20}},
\code{\link{madat21}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat5}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}},
\code{\link{moregood}},
\code{\link{mostluck}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{noemote}},
\code{\link{nogood}},
\code{\link{noplan}},
\code{\link{notcount}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{othresp}},
\code{\link{outraged}},
\code{\link{ovrjoyed}},
\code{\link{pessimst}},
\code{\link{proud}},
\code{\link{reactok}},
\code{\link{restless}},
\code{\link{sad}},
\code{\link{satself}},
\code{\link{selfless}},
\code{\link{selfresp}},
\code{\link{shakeblu}},
\code{\link{showangr}},
\code{\link{showfeel}},
\code{\link{slfrspct}},
\code{\link{whnangry}},
\code{\link{worried}},
\code{\link{wrthless}}
}
\concept{Emotions}
\concept{Feelings}
\keyword{variable}
