% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{threaten}
\alias{threaten}
\title{R is threatened or harassed}
\source{
General Social Survey https://gss.norc.org
}
\description{
threaten
}
\details{
Question In your day-to-day life how often have any of the following things happened to you? You are threatened or harassed.
}
\section{Values}{

\itemize{
\item \code{1} almost every day
\item \code{2} at least once a week
\item \code{3} a few times a month
\item \code{4} a few times a year
\item \code{5} less than once a year
\item \code{6} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7162/vshow}{GSS Data Explorer page for \code{threaten}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab almost every day \tab at least once a week \tab a few times a month \tab a few times a year \tab less than once a year \tab never \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab 774 \tab 14 \tab 30 \tab 39 \tab 155 \tab 344 \tab 982 \tab 5 \tab 5 \tab - \tab 2348 \cr
   2021 \tab 1411 \tab 20 \tab 44 \tab 84 \tab 322 \tab 738 \tab 1396 \tab - \tab - \tab 17 \tab 4032 \cr
   2022 \tab 1207 \tab 28 \tab 54 \tab 94 \tab 260 \tab 591 \tab 1296 \tab 7 \tab - \tab 7 \tab 3544 \cr
   Total \tab 3392 \tab 62 \tab 128 \tab 217 \tab 737 \tab 1673 \tab 3674 \tab 12 \tab 5 \tab 24 \tab 9924 \cr
}
}

\keyword{variable}
