% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{timefam}
\alias{timefam}
\title{Time r would spend on family}
\source{
General Social Survey https://gss.norc.org
}
\description{
timefam
}
\details{
Question Suppose you could change the way you spend your time, spending more on some things and less on others. Which of these things on the following list would you like to spend more time on, which would you like to spend less time on, and which would you like to spend the same amount of time as now?
C. Time with your family?
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend a bit more
\item \code{3} spend same
\item \code{4} spend a bit less
\item \code{5} spend much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4488/vshow}{GSS Data Explorer page for \code{timefam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend a bit more \tab spend same \tab spend a bit less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 467 \tab 588 \tab 311 \tab 26 \tab 10 \tab 15 \tab 36 \tab 1537 \cr
   1998 \tab 1609 \tab 490 \tab 447 \tab 213 \tab 7 \tab 12 \tab 21 \tab 33 \tab 2832 \cr
   2006 \tab 2992 \tab 839 \tab 406 \tab 236 \tab 13 \tab 6 \tab 16 \tab 2 \tab 4510 \cr
   Total \tab 4685 \tab 1796 \tab 1441 \tab 760 \tab 46 \tab 28 \tab 52 \tab 71 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timepdwk}},
\code{\link{timerelx}}
}
\concept{Work Orientation}
\keyword{variable}
