% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{timefrnd}
\alias{timefrnd}
\title{Time r would spend on friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
timefrnd
}
\details{
Question Suppose you could change the way you spend your time, spending more on some things and less on others. Which of these things on the following list would you like to spend more time on, which would you like to spend less time on, and which would you like to spend the same amount of time as now?
D. Time with your friends?
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend a bit more
\item \code{3} spend same
\item \code{4} spend a bit less
\item \code{5} spend much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4489/vshow}{GSS Data Explorer page for \code{timefrnd}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab spend much more \tab spend a bit more \tab spend same \tab spend a bit less \tab spend much less \tab can't choose \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 240 \tab 644 \tab 462 \tab 38 \tab 14 \tab 18 \tab 121 \tab - \tab 1537 \cr
   1998 \tab - \tab 230 \tab 536 \tab 361 \tab 29 \tab 15 \tab 14 \tab 202 \tab 1445 \tab 2832 \cr
   2006 \tab - \tab 313 \tab 540 \tab 554 \tab 60 \tab 29 \tab 20 \tab 2 \tab 2992 \tab 4510 \cr
   Total \tab 0 \tab 783 \tab 1720 \tab 1377 \tab 127 \tab 58 \tab 52 \tab 325 \tab 4437 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{timefam}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timepdwk}},
\code{\link{timerelx}}
}
\concept{Work Orientation}
\keyword{variable}
