% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{timehhwk}
\alias{timehhwk}
\title{Time r would spend on household work}
\source{
General Social Survey https://gss.norc.org
}
\description{
timehhwk
}
\details{
Question Suppose you could change the way you spend your time, spending more on some things and less on others. Which of these things on the following list would you like to spend more time on, which would you like to spend less time on, and which would you like to spend the same amount of time as now?
B. Time doing household work?
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend a bit more
\item \code{3} spend same
\item \code{4} spend a bit less
\item \code{5} spend much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4487/vshow}{GSS Data Explorer page for \code{timehhwk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend a bit more \tab spend same \tab spend a bit less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 51 \tab 284 \tab 534 \tab 288 \tab 227 \tab 26 \tab 43 \tab 1537 \cr
   1998 \tab 1609 \tab 57 \tab 233 \tab 385 \tab 247 \tab 231 \tab 30 \tab 40 \tab 2832 \cr
   2006 \tab 2992 \tab 98 \tab 297 \tab 532 \tab 257 \tab 282 \tab 49 \tab 3 \tab 4510 \cr
   Total \tab 4685 \tab 206 \tab 814 \tab 1451 \tab 792 \tab 740 \tab 105 \tab 86 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timeleis}},
\code{\link{timepdwk}},
\code{\link{timerelx}}
}
\concept{Work Orientation}
\keyword{variable}
