% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{timekid1}
\alias{timekid1}
\title{Time r spends with child 1}
\source{
General Social Survey https://gss.norc.org
}
\description{
timekid1
}
\details{
Question Would you say that the amount of time you spend with (NAME OF CHILD) is too much, about the right amount, or too little?
A. First listed child?
}
\section{Values}{

\itemize{
\item \verb{[1]} too much
\item \verb{[2]} about the right amount
\item \verb{[3]} too little
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2639/vshow}{GSS Data Explorer page for \code{timekid1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too much \tab about the right amount \tab too little \tab no answer \tab not available in this year \tab Total \cr
   2000 \tab 1734 \tab 47 \tab 482 \tab 425 \tab 129 \tab - \tab 2817 \cr
   Total \tab 1734 \tab 47 \tab 482 \tab 425 \tab 129 \tab 0 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artgn}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyinfgn}},
\code{\link{buyit12}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{gamegn}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{huclean}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{inperson}},
\code{\link{intaccss}},
\code{\link{invest12}},
\code{\link{investgn}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobhrw}},
\code{\link{jobinfo}},
\code{\link{jobminw}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{letters}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{meetings}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{newjobgn}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numcntct}},
\code{\link{numhome}},
\code{\link{numprobs}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{peoplegn}},
\code{\link{perfrmgn}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polcamp}},
\code{\link{polcangn}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldisgn}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polfrel}},
\code{\link{polinf12}},
\code{\link{polinfgn}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{polpapr}},
\code{\link{poltough}},
\code{\link{poltv}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{travelgn}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}

Other Children: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adoption}},
\code{\link{adoptkid}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{aidkids}},
\code{\link{allergic}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{cathsch}},
\code{\link{chembal}},
\code{\link{childs}},
\code{\link{chldben}},
\code{\link{chldcare}},
\code{\link{chldeduc}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldhrt}},
\code{\link{chldidel}},
\code{\link{chldidel_next}},
\code{\link{chldlove}},
\code{\link{chldmore}},
\code{\link{chldmorl}},
\code{\link{chldnum}},
\code{\link{chldpref_next}},
\code{\link{chldprob}},
\code{\link{chldsafe}},
\code{\link{chldsch}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldskls}},
\code{\link{chldsoon}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{chldvig}},
\code{\link{chrissch}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{comknows}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{discipln}},
\code{\link{eldersup}},
\code{\link{evkid}},
\code{\link{failure}},
\code{\link{famper4}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{fourkids}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{helpoth}},
\code{\link{hlth8}},
\code{\link{homesch}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{illphys}},
\code{\link{impkids}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpicked}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kid5up}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidhlth}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidlived}},
\code{\link{kidneedy}},
\code{\link{kidnofre}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidnum}},
\code{\link{kidout}},
\code{\link{kidpars}},
\code{\link{kidresp}},
\code{\link{kidskips}},
\code{\link{kidsocst}},
\code{\link{kidssol}},
\code{\link{kidtrble}},
\code{\link{kidxfilm}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{law7}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{mapaid}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{mntlill}},
\code{\link{morextrm}},
\code{\link{mykids}},
\code{\link{nextdoor}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obeythnk}},
\code{\link{obrespct}},
\code{\link{onekid}},
\code{\link{oth3}},
\code{\link{othkids}},
\code{\link{othsch}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{popular}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{raised}},
\code{\link{relsch}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{sellbaby}},
\code{\link{spanking}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spendeve}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}},
\code{\link{spwrkpar}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{stressfl}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{sufadult}},
\code{\link{thnkself}},
\code{\link{threkids}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{trbllaw}},
\code{\link{twokids}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{workhard}},
\code{\link{zombies}}
}
\concept{Children}
\concept{Information Society}
\keyword{variable}
