% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{timepdwk}
\alias{timepdwk}
\title{Time r would spend on paid work}
\source{
General Social Survey https://gss.norc.org
}
\description{
timepdwk
}
\details{
Question Suppose you could change the way you spend your time, spending more on some things and less on others. Which of these things on the following list would you like to spend more time on, which would you like to spend less time on, and which would you like to spend the same amount of time as now?
A. Time in a paid job?
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend a bit more
\item \verb{[3]} spend same
\item \verb{[4]} spend a bit less
\item \verb{[5]} spend much less
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4486/vshow}{GSS Data Explorer page for \code{timepdwk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend a bit more \tab spend same \tab spend a bit less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 114 \tab 186 \tab 570 \tab 263 \tab 122 \tab 146 \tab 52 \tab 1537 \cr
   1998 \tab 1609 \tab 110 \tab 151 \tab 350 \tab 204 \tab 105 \tab 227 \tab 76 \tab 2832 \cr
   2006 \tab 2992 \tab 140 \tab 162 \tab 578 \tab 283 \tab 115 \tab 237 \tab 3 \tab 4510 \cr
   Total \tab 4685 \tab 364 \tab 499 \tab 1498 \tab 750 \tab 342 \tab 610 \tab 131 \tab 8879 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timerelx}}
}
\concept{Work Orientation}
\keyword{variable}
