% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_487.R
\name{tiredwrk}
\alias{tiredwrk}
\title{Arrived at work too tired to function bc of hhwork}
\source{
General Social Survey https://gss.norc.org
}
\description{
tiredwrk
}
\details{
Question How often has each of the following happened to you during the past three months?
C. I have arrived at work too tired to function well because of the household work I had done.
}
\section{Values}{

\itemize{
\item \verb{[1]} several times a week
\item \verb{[2]} several times a month
\item \verb{[3]} once or twice
\item \verb{[4]} never
\item \verb{[6]} doesnt apply or no job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4449/vshow}{GSS Data Explorer page for \code{tiredwrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab several times a week \tab several times a month \tab once or twice \tab never \tab doesnt apply or no job \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1594 \tab 27 \tab 48 \tab 244 \tab 524 \tab 318 \tab 4 \tab 6 \tab 2765 \cr
   Total \tab 1594 \tab 27 \tab 48 \tab 244 \tab 524 \tab 318 \tab 4 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{buythngs}},
\code{\link{famvswrk}},
\code{\link{jobvsfam}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{weekend}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
