% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{tithing}
\alias{tithing}
\title{How much does r give to religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
tithing
}
\details{
Question About how much do you contribute to your religion every year (not including school tuition)?
}
\section{Values}{

\itemize{
\item \code{99995} tithes, amt. not given
\item \code{99996} 99996 or more
\item \code{99997} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/360/vshow}{GSS Data Explorer page for \code{tithing}}.
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}}
}
\concept{Core}
\keyword{variable}
