% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{tlkfam}
\alias{tlkfam}
\title{X should talk to family and friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
tlkfam
}
\details{
Question Should NAME do any of the following:
A. Talk things over with family and friends
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2011/vshow}{GSS Data Explorer page for \code{tlkfam}}.
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
