% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_432.R
\name{toldexp2}
\alias{toldexp2}
\title{R has ever told anyone about experience with leader2}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldexp2
}
\details{
Question B. Have you ever told anyone about this experience?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3961/vshow}{GSS Data Explorer page for \code{toldexp2}}.
}

\seealso{
Other Clergy/Sex: 
\code{\link{clsclsex}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{knwclsex}},
\code{\link{toldexp1}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
