% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_432.R
\name{toldexp3}
\alias{toldexp3}
\title{R has ever told anyone about experience with leader3}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldexp3
}
\details{
Question C. Have you ever told anyone about this experience?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3962/vshow}{GSS Data Explorer page for \code{toldexp3}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 2021 \tab 1 \tab 1 \tab 2023 \cr
   Total \tab 2021 \tab 1 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clsclsex}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{knwclsex}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldrel1}},
\code{\link{toldrel2}},
\code{\link{toldrel3}}
}
\concept{Clergy/Sex}
\keyword{variable}
