% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_432.R
\name{toldrel1}
\alias{toldrel1}
\title{R ever told any authority in congregation about experience with leader1}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldrel1
}
\details{
Question A. Did you ever tell anyone in a position of authority in your congregation, denomination, or religious group  about this experience?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3963/vshow}{GSS Data Explorer page for \code{toldrel1}}.
}

\seealso{
Other Clergy/Sex: 
\code{\link{clsclsex}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{knwclsex}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel2}},
\code{\link{toldrel3}}

Other Religion: 
\code{\link{clsclsex}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{knwclsex}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel2}},
\code{\link{toldrel3}}

Other Sex: 
\code{\link{clsclsex}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{knwclsex}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel2}},
\code{\link{toldrel3}}
}
\concept{Clergy/Sex}
\concept{Religion}
\concept{Sex}
\keyword{variable}
