% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_432.R
\name{toldrel3}
\alias{toldrel3}
\title{R ever told any authority in congregation about experience with leader3}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldrel3
}
\details{
Question C. Did you ever tell anyone in a position of authority in your congregation, denomination, or religious group  about this experience?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3965/vshow}{GSS Data Explorer page for \code{toldrel3}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab no \tab Total \cr
   2008 \tab 2022 \tab 1 \tab 2023 \cr
   Total \tab 2022 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{clsclsex}},
\code{\link{hushrel2}},
\code{\link{hushrel3}},
\code{\link{knwclsex}},
\code{\link{toldexp1}},
\code{\link{toldexp2}},
\code{\link{toldexp3}},
\code{\link{toldrel1}},
\code{\link{toldrel2}}
}
\concept{Clergy/Sex}
\keyword{variable}
