% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{toldsmsx}
\alias{toldsmsx}
\title{At what age you first told someone that you had sex with same sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldsmsx
}
\details{
Question At about what age were you when you first told someone that you had sex with someone of the same sex?
}
\section{Values}{

\itemize{
\item \verb{[1]} under age 14
\item \verb{[2]} age 14-18
\item \verb{[3]} age 18-25
\item \verb{[4]} age 26-35
\item \verb{[5]} age 36-45
\item \verb{[6]} age 46-55
\item \verb{[7]} age 56-65
\item \verb{[8]} age 66 or older
\item \verb{[9]} i have never told anyone
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3907/vshow}{GSS Data Explorer page for \code{toldsmsx}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab under age 14 \tab age 14-18 \tab age 18-25 \tab age 26-35 \tab age 36-45 \tab age 46-55 \tab i have never told anyone \tab Total \cr
   2008 \tab 1951 \tab 7 \tab 14 \tab 26 \tab 12 \tab 2 \tab 3 \tab 8 \tab 2023 \cr
   Total \tab 1951 \tab 7 \tab 14 \tab 26 \tab 12 \tab 2 \tab 3 \tab 8 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{sexsex18}}
}
\concept{Sexual Orientation}
\keyword{variable}
