% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{toldsxor}
\alias{toldsxor}
\title{At what age you first told someone that you were gay/les/bi}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldsxor
}
\details{
Question At about what age were you when you first told someone that you were (gay(IF MALE)/lesbian(IF FEMALE)/bisexual)?
}
\section{Values}{

\itemize{
\item \code{1} under age 14
\item \code{2} age 14-18
\item \code{3} age 18-25
\item \code{4} age 26-35
\item \code{5} age 36-45
\item \code{6} age 46-55
\item \code{7} age 56-65
\item \code{8} age 66 or older
\item \code{9} i have never told anyone
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3909/vshow}{GSS Data Explorer page for \code{toldsxor}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab under age 14 \tab age 14-18 \tab age 18-25 \tab age 26-35 \tab age 36-45 \tab age 56-65 \tab i have never told anyone \tab no answer \tab Total \cr
   2008 \tab 1942 \tab 3 \tab 10 \tab 20 \tab 10 \tab 3 \tab 3 \tab 9 \tab 23 \tab 2023 \cr
   Total \tab 1942 \tab 3 \tab 10 \tab 20 \tab 10 \tab 3 \tab 3 \tab 9 \tab 23 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{sexunion}}
}
\concept{Sexual Orientation}
\keyword{variable}
