% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{toldwork}
\alias{toldwork}
\title{How many of r's coworkers know r is gay/les/bi}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldwork
}
\details{
Question Some people are very open about their sexual orientation while others are not. About how many  of your coworkers know that you (are gay(IF MALE)/are lesbian(IF FEMALE)/are bisexual/have sex with men(IF MALE)/have sex with women(IF FEMALE))?
}
\section{Values}{

\itemize{
\item \code{1} all of my coworkers
\item \code{2} more than half
\item \code{3} about half
\item \code{4} less than half
\item \code{5} none of them
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3919/vshow}{GSS Data Explorer page for \code{toldwork}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}}

Other Gays And Lesbians: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{losejob5}},
\code{\link{negjob5}}

Other Job: 
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{harjob5}},
\code{\link{negjob5}}

Other Work: 
\code{\link{evharjb}},
\code{\link{harjob5}},
\code{\link{negjob5}}
}
\concept{Gays And Lesbians}
\concept{Job}
\concept{Sexual Orientation}
\concept{Work}
\keyword{variable}
