% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{tomatoes}
\alias{tomatoes}
\title{Regular tomatoes no genes}
\source{
General Social Survey https://gss.norc.org
}
\description{
tomatoes
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don’t know or aren’t sure, just tell me so,  and we will skip to the next question. Remember true, false, or don’t know.
L. Ordinary tomatoes do not contain genes, while genetically modified tomatoes Do.
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3345/vshow}{GSS Data Explorer page for \code{tomatoes}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}}
}
\concept{Genetic Testing}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
