% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{toodifme}
\alias{toodifme}
\title{Too diff to do anything about envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
toodifme
}
\details{
Question How much do you agree or disagree with each of these statements?
A. It is just too difficult for someone like me to do much about the environment.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4739/vshow}{GSS Data Explorer page for \code{toodifme}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 69 \tab 305 \tab 204 \tab 748 \tab 129 \tab 21 \tab 81 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 63 \tab 295 \tab 232 \tab 629 \tab 109 \tab 18 \tab 40 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 71 \tab 269 \tab 241 \tab 516 \tab 92 \tab 21 \tab 66 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 88 \tab 340 \tab 214 \tab 613 \tab 127 \tab 40 \tab 8 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 62 \tab 305 \tab 474 \tab 756 \tab 186 \tab 60 \tab 178 \tab 8 \tab 4032 \cr
   Total \tab 5813 \tab 353 \tab 1514 \tab 1365 \tab 3262 \tab 643 \tab 160 \tab 373 \tab 8 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}}
}
\concept{ISSP Environment}
\keyword{variable}
