% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{toofast}
\alias{toofast}
\title{Science makes our way of life change too fast}
\source{
General Social Survey https://gss.norc.org
}
\description{
toofast
}
\details{
Question I’m going to read to you some statements like those you might find in a newspaper or magazine article. For each statement, please tell me if you strongly agree, agree, disagree, or strongly disagree.
B. Science makes our way of life change too fast.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008, 2022 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3318/vshow}{GSS Data Explorer page for \code{toofast}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2006 \tab 2646 \tab 172 \tab 662 \tab 824 \tab 153 \tab 52 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 518 \tab 137 \tab 560 \tab 646 \tab 112 \tab 47 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 1353 \tab 53 \tab 272 \tab 291 \tab 54 \tab 17 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 92 \tab 315 \tab 512 \tab 40 \tab 37 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 158 \tab 485 \tab 506 \tab 70 \tab 18 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 162 \tab 552 \tab 541 \tab 106 \tab 25 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 159 \tab 418 \tab 490 \tab 93 \tab 13 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 2150 \tab 163 \tab 617 \tab 875 \tab 207 \tab 2 \tab - \tab 18 \tab 4032 \cr
   2022 \tab 2309 \tab 108 \tab 443 \tab 553 \tab 129 \tab - \tab - \tab 2 \tab 3544 \cr
   Total \tab 13897 \tab 1204 \tab 4324 \tab 5238 \tab 964 \tab 211 \tab 22 \tab 20 \tab 25880 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{whichsci}}

Other GSS Next: 
\code{\link{advfront}},
\code{\link{nextgen}}
}
\concept{GSS Next}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
