% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{topprob1_next}
\alias{topprob1_next}
\title{Most important issue for america - follow on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
topprob1_next
}
\details{
Question What do you think is the most important issue for America today?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} health care
\item \verb{[2]} education
\item \verb{[3]} crime
\item \verb{[4]} the environment
\item \verb{[5]} immigration
\item \verb{[6]} the economy
\item \verb{[7]} terrorism
\item \verb{[8]} poverty
\item \verb{[9]} none of these
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8292/vshow}{GSS Data Explorer page for \code{topprob1_next}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab health care \tab education \tab crime \tab the environment \tab immigration \tab the economy \tab terrorism \tab poverty \tab none of these \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 127 \tab 60 \tab 73 \tab 35 \tab 87 \tab 268 \tab 31 \tab 65 \tab 286 \tab 2262 \tab 15 \tab 3309 \cr
   Total \tab 0 \tab 127 \tab 60 \tab 73 \tab 35 \tab 87 \tab 268 \tab 31 \tab 65 \tab 286 \tab 2262 \tab 15 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{happy_next}},
\code{\link{knwsexor}},
\code{\link{perla_self}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other Crime: 
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other Economy: 
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other Education: 
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other Environment: 
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other Foreign Affairs: 
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other Salience Of Issues: 
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}

Other United States: 
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}}
}
\concept{Crime}
\concept{Economy}
\concept{Education}
\concept{Environment}
\concept{Foreign Affairs}
\concept{GSS Next}
\concept{Salience Of Issues}
\concept{United States}
\keyword{variable}
