% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{topprob2_envr}
\alias{topprob2_envr}
\title{Next important issue (environment) - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
topprob2_envr
}
\details{
Question Do any other issues match your answer to the previous question (TOPPROB1_NEXT)? The environment
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} mentioned
\item \verb{[2]} not mentioned
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8278/vshow}{GSS Data Explorer page for \code{topprob2_envr}}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab mentioned \tab not mentioned \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 173 \tab 642 \tab 2261 \tab 233 \tab 3309 \cr
   Total \tab 0 \tab 173 \tab 642 \tab 2261 \tab 233 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{happy_next}},
\code{\link{knwsexor}},
\code{\link{perla_self}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}

Other Crime: 
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}

Other Economy: 
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}

Other Education: 
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}

Other Environment: 
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}

Other Foreign Affairs: 
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}

Other Salience Of Issues: 
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}

Other United States: 
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_hlth}}
}
\concept{Crime}
\concept{Economy}
\concept{Education}
\concept{Environment}
\concept{Foreign Affairs}
\concept{GSS Next}
\concept{Salience Of Issues}
\concept{United States}
\keyword{variable}
