% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{totalnum}
\alias{totalnum}
\title{Number of employees: r's entire firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
totalnum
}
\details{
Question About how many people are employed in the entire business, corporation or organization for which you work? (Count  part-time as well as full-time employees. (If you work for the government, how many people are employed in your  whole division or agency?)
}
\section{Values}{

\itemize{
\item \verb{[1]} 1-9
\item \verb{[2]} 10-49
\item \verb{[3]} 50-99
\item \verb{[4]} 100-499
\item \verb{[5]} 500-999
\item \verb{[6]} 1,000-1,999
\item \verb{[7]} 2,000 - 9,999
\item \verb{[8]} 10,000+
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4568/vshow}{GSS Data Explorer page for \code{totalnum}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1-9 \tab 10-49 \tab 50-99 \tab 100-499 \tab 500-999 \tab 1,000-1,999 \tab 2,000 - 9,999 \tab 10,000+ \tab don't know \tab Total \cr
   1989 \tab 709 \tab 81 \tab 124 \tab 58 \tab 135 \tab 50 \tab 38 \tab 83 \tab 103 \tab 156 \tab 1537 \cr
   Total \tab 709 \tab 81 \tab 124 \tab 58 \tab 135 \tab 50 \tab 38 \tab 83 \tab 103 \tab 156 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}}
}
\concept{Work Orientation}
\keyword{variable}
