% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_206.R
\name{toteduc}
\alias{toteduc}
\title{How much money contributed in education}
\source{
General Social Survey https://gss.norc.org
}
\description{
toteduc
}
\details{
Question Approximately how much money or the cash equivalent of property have you contributed in each of the fields listed above in the past twelve months? (REPEAT FOR EACH ITEM MENTIONED IN GIVHLTH-GIVOTH) RECORD ABOVE UNDER TOTHLTH-VALOTH. (ESTIMATED DOLLAR VALUE CONTRIBUTED INCLUDING BOTH CASH CONTRIBUTIONS AND THE CASH-VALUE OF PROPERTY.)
B. Education
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1946/vshow}{GSS Data Explorer page for \code{toteduc}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 10000 \tab 1100 \tab 120 \tab 1200 \tab 125 \tab 130 \tab 140 \tab 15 \tab 150 \tab 16 \tab 180 \tab 20 \tab 200 \tab 2000 \tab 225 \tab 25 \tab 250 \tab 2500 \tab 30 \tab 300 \tab 35 \tab 350 \tab 36 \tab 40 \tab 400 \tab 4000 \tab 5 \tab 50 \tab 500 \tab 6 \tab 60 \tab 600 \tab 6000 \tab 7000 \tab 80 \tab 90 \tab 99996 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2685 \tab 3 \tab 4 \tab 24 \tab 6 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 11 \tab 24 \tab 4 \tab 1 \tab 14 \tab 3 \tab 3 \tab 5 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 4 \tab 27 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 22 \tab 2904 \cr
   Total \tab 2685 \tab 3 \tab 4 \tab 24 \tab 6 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 11 \tab 24 \tab 4 \tab 1 \tab 14 \tab 3 \tab 3 \tab 5 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 4 \tab 27 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 22 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{givinfrm}},
\code{\link{givintl}},
\code{\link{givoth}},
\code{\link{totenvir}},
\code{\link{tothlth}},
\code{\link{tothuman}},
\code{\link{totpub}},
\code{\link{totrec}},
\code{\link{totrelig}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
