% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_207.R
\name{totintl}
\alias{totintl}
\title{How much money contributed in international}
\source{
General Social Survey https://gss.norc.org
}
\description{
totintl
}
\details{
Question Approximately how much money or the cash equivalent of property have you contributed in each of the fields listed above in the past twelve months? (REPEAT FOR EACH ITEM MENTIONED IN GIVHLTH-GIVOTH) RECORD ABOVE UNDER TOTHLTH-VALOTH. (ESTIMATED DOLLAR VALUE CONTRIBUTED INCLUDING BOTH CASH CONTRIBUTIONS AND THE CASH-VALUE OF PROPERTY.)
M. International/foreign
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1957/vshow}{GSS Data Explorer page for \code{totintl}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 10 \tab 100 \tab 1000 \tab 15 \tab 200 \tab 2000 \tab 250 \tab 300 \tab 36 \tab 40 \tab 50 \tab 500 \tab 60 \tab 75 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2871 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 6 \tab 2904 \cr
   Total \tab 2871 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{totart}},
\code{\link{totfound}},
\code{\link{totinfrm}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totwork}},
\code{\link{totyouth}},
\code{\link{valeduc}},
\code{\link{valhlth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
