% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_206.R
\name{totrelig}
\alias{totrelig}
\title{How much money contributed in religious org}
\source{
General Social Survey https://gss.norc.org
}
\description{
totrelig
}
\details{
Question Approximately how much money or the cash equivalent of property have you contributed in each of the fields listed above in the past twelve months? (REPEAT FOR EACH ITEM MENTIONED IN GIVHLTH-GIVOTH) RECORD ABOVE UNDER TOTHLTH-VALOTH. (ESTIMATED DOLLAR VALUE CONTRIBUTED INCLUDING BOTH CASH CONTRIBUTIONS AND THE CASH-VALUE OF PROPERTY.)
C. Religious organizations
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1947/vshow}{GSS Data Explorer page for \code{totrelig}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 10000 \tab 10400 \tab 1100 \tab 11000 \tab 120 \tab 1200 \tab 130 \tab 1300 \tab 140 \tab 1400 \tab 15 \tab 150 \tab 1500 \tab 1600 \tab 1700 \tab 1750 \tab 20 \tab 200 \tab 2000 \tab 2200 \tab 2300 \tab 240 \tab 2400 \tab 25 \tab 250 \tab 2500 \tab 265 \tab 2800 \tab 2952 \tab 30 \tab 300 \tab 3000 \tab 3200 \tab 3400 \tab 35 \tab 350 \tab 3500 \tab 35000 \tab 360 \tab 4 \tab 40 \tab 400 \tab 4000 \tab 40000 \tab 450 \tab 4500 \tab 460 \tab 48 \tab 480 \tab 492 \tab 5 \tab 50 \tab 500 \tab 5000 \tab 520 \tab 5200 \tab 530 \tab 5385 \tab 55 \tab 550 \tab 5500 \tab 60 \tab 600 \tab 6000 \tab 625 \tab 70 \tab 700 \tab 7000 \tab 75 \tab 80 \tab 800 \tab 8000 \tab 85 \tab 900 \tab 9000 \tab 99996 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2272 \tab 6 \tab 4 \tab 51 \tab 31 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 12 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 16 \tab 11 \tab 3 \tab 1 \tab 1 \tab 17 \tab 36 \tab 17 \tab 1 \tab 1 \tab 3 \tab 3 \tab 15 \tab 13 \tab 10 \tab 1 \tab 1 \tab 1 \tab 9 \tab 37 \tab 9 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 21 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 37 \tab 43 \tab 8 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 16 \tab 6 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 43 \tab 62 \tab 2904 \cr
   Total \tab 2272 \tab 6 \tab 4 \tab 51 \tab 31 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 12 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 16 \tab 11 \tab 3 \tab 1 \tab 1 \tab 17 \tab 36 \tab 17 \tab 1 \tab 1 \tab 3 \tab 3 \tab 15 \tab 13 \tab 10 \tab 1 \tab 1 \tab 1 \tab 9 \tab 37 \tab 9 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 21 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 37 \tab 43 \tab 8 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 16 \tab 6 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 43 \tab 62 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{givinfrm}},
\code{\link{givintl}},
\code{\link{givoth}},
\code{\link{toteduc}},
\code{\link{totenvir}},
\code{\link{tothlth}},
\code{\link{tothuman}},
\code{\link{totpub}},
\code{\link{totrec}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
