% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{totsup}
\alias{totsup}
\title{How many people responsible to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
totsup
}
\details{
Question Counting all of these levels, how may people are responsible to you both directly and indirectly?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1317/vshow}{GSS Data Explorer page for \code{totsup}}.
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{yousup}}

Other Job: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{yousup}}

Other Work: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{yousup}}
}
\concept{Job}
\concept{Work}
\concept{Work Organizations}
\keyword{variable}
