% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{totunemp}
\alias{totunemp}
\title{Total times unemployed}
\source{
General Social Survey https://gss.norc.org
}
\description{
totunemp
}
\details{
Question Counting all these times, in total were you unemployed and looking for work for…
}
\section{Values}{

\itemize{
\item \verb{[1]} 2-8 weeks
\item \verb{[2]} 8 week - 6 mon
\item \verb{[3]} 6-12 mon
\item \verb{[4]} ) 1 year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1274/vshow}{GSS Data Explorer page for \code{totunemp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 2-8 weeks \tab 8 week - 6 mon \tab 6-12 mon \tab ) 1 year \tab don't know \tab no answer \tab Total \cr
   1991 \tab 824 \tab 337 \tab 187 \tab 75 \tab 81 \tab 2 \tab 11 \tab 1517 \cr
   Total \tab 824 \tab 337 \tab 187 \tab 75 \tab 81 \tab 2 \tab 11 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{numunemp}},
\code{\link{orgfin}},
\code{\link{wrkyears}}
}
\concept{Work Organizations}
\keyword{variable}
