% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_207.R
\name{totwork}
\alias{totwork}
\title{How much money contributed in work-related org}
\source{
General Social Survey https://gss.norc.org
}
\description{
totwork
}
\details{
Question Approximately how much money or the cash equivalent of property have you contributed in each of the fields listed above in the past twelve months? (REPEAT FOR EACH ITEM MENTIONED IN GIVHLTH-GIVOTH) RECORD ABOVE UNDER TOTHLTH-VALOTH. (ESTIMATED DOLLAR VALUE CONTRIBUTED INCLUDING BOTH CASH CONTRIBUTIONS AND THE CASH-VALUE OF PROPERTY.)
I. Work-related organizations
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1953/vshow}{GSS Data Explorer page for \code{totwork}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 120 \tab 15 \tab 150 \tab 20 \tab 200 \tab 2000 \tab 25 \tab 26 \tab 28 \tab 30 \tab 300 \tab 3000 \tab 35 \tab 350 \tab 40 \tab 5 \tab 50 \tab 500 \tab 60 \tab 600 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2800 \tab 1 \tab 4 \tab 20 \tab 1 \tab 1 \tab 3 \tab 4 \tab 9 \tab 6 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 9 \tab 4 \tab 2 \tab 1 \tab 10 \tab 6 \tab 2904 \cr
   Total \tab 2800 \tab 1 \tab 4 \tab 20 \tab 1 \tab 1 \tab 3 \tab 4 \tab 9 \tab 6 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 9 \tab 4 \tab 2 \tab 1 \tab 10 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{totart}},
\code{\link{totfound}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totyouth}},
\code{\link{valeduc}},
\code{\link{valhlth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
