% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{trainops}
\alias{trainops}
\title{R have the training opportunities}
\source{
General Social Survey https://gss.norc.org
}
\description{
trainops
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
O. I have the training opportunities I need to perform my job safely and competently
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not too true
\item \verb{[4]} not at all true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2810/vshow}{GSS Data Explorer page for \code{trainops}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 1125 \tab 504 \tab 78 \tab 58 \tab 15 \tab 16 \tab 2765 \cr
   2006 \tab 2776 \tab 1011 \tab 537 \tab 89 \tab 50 \tab 25 \tab 22 \tab 4510 \cr
   Total \tab 3745 \tab 2136 \tab 1041 \tab 167 \tab 108 \tab 40 \tab 38 \tab 7275 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
