% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{trainsch}
\alias{trainsch}
\title{How many years spent on training}
\source{
General Social Survey https://gss.norc.org
}
\description{
trainsch
}
\details{
Question B. In total, how many years have you spent (on such specialized training in school)?
}
\section{Values}{

\itemize{
\item \verb{[1]} less than one year
\item \verb{[2]} two years or less
\item \verb{[3]} four years or less
\item \verb{[4]} six years or less
\item \verb{[5]} eight years or less
\item \verb{[6]} over eight years
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2139/vshow}{GSS Data Explorer page for \code{trainsch}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab less than one year \tab two years or less \tab four years or less \tab six years or less \tab eight years or less \tab over eight years \tab no answer \tab Total \cr
   1998 \tab 2194 \tab 136 \tab 150 \tab 110 \tab 50 \tab 19 \tab 29 \tab 144 \tab 2832 \cr
   Total \tab 2194 \tab 136 \tab 150 \tab 110 \tab 50 \tab 19 \tab 29 \tab 144 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{typeorg}}
}
\concept{Jobs}
\keyword{variable}
