% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_197.R
\name{transoth}
\alias{transoth}
\title{Hwould you be to mention problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
transoth
}
\details{
Question Now imagine that someone you didn't know comes to see the car and wants to buy it. How likely would you be to mention the car's history of transmission problems (if this person did not ask)?
}
\section{Values}{

\itemize{
\item \code{1} definitely would not mention it
\item \code{2} probably would not
\item \code{3} probably would
\item \code{4} definitely would mention it
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1854/vshow}{GSS Data Explorer page for \code{transoth}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely would not mention it \tab probably would not \tab probably would \tab definitely would mention it \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2168 \tab 75 \tab 120 \tab 242 \tab 280 \tab 14 \tab 5 \tab 2904 \cr
   Total \tab 2168 \tab 75 \tab 120 \tab 242 \tab 280 \tab 14 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{sellauto}},
\code{\link{sellhome}}
}
\concept{Markets}
\keyword{variable}
