% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{transrel}
\alias{transrel}
\title{How likely to mention problems to relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
transrel
}
\details{
Question Now imagine that one of your relatives sees the car and wants to buy it. How likely would you be to mention the car's history of transmission problems (if this person did not ask)?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely would not mention it
\item \verb{[2]} probably would not
\item \verb{[3]} probably would
\item \verb{[4]} definitely would mention it
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1862/vshow}{GSS Data Explorer page for \code{transrel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely would not mention it \tab probably would not \tab probably would \tab definitely would mention it \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2196 \tab 71 \tab 27 \tab 109 \tab 492 \tab 6 \tab 3 \tab 2904 \cr
   Total \tab 2196 \tab 71 \tab 27 \tab 109 \tab 492 \tab 6 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}}
}
\concept{Markets}
\keyword{variable}
