% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_84.R
\name{trarel1}
\alias{trarel1}
\title{Trauma of relatives, last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
trarel1
}
\details{
Question A. Number of traumatic events (divorces, unemployment, hospitalizations/disabilities) happening to respondent's relatives last year.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/755/vshow}{GSS Data Explorer page for \code{trarel1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab Total \cr
   1978 \tab - \tab 614 \tab 608 \tab 264 \tab 46 \tab 1532 \cr
   1980 \tab - \tab 601 \tab 587 \tab 238 \tab 42 \tab 1468 \cr
   1983 \tab - \tab 559 \tab 645 \tab 338 \tab 57 \tab 1599 \cr
   1984 \tab - \tab 584 \tab 596 \tab 254 \tab 39 \tab 1473 \cr
   Total \tab 0 \tab 2358 \tab 2436 \tab 1094 \tab 184 \tab 6072 \cr
}
}

\seealso{
Other Core: 
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{sibdeath}},
\code{\link{spdeath}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}
}
\concept{Core}
\keyword{variable}
