% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_84.R
\name{trarel5}
\alias{trarel5}
\title{Trauma of relatives, last 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
trarel5
}
\details{
Question B. Number of traumatic events (divorces, unemployment, hospitalizations/disabilities) happening to respondent''s relatives during last five years.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/756/vshow}{GSS Data Explorer page for \code{trarel5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab Total \cr
   1978 \tab - \tab 280 \tab 573 \tab 483 \tab 196 \tab 1532 \cr
   1980 \tab - \tab 282 \tab 554 \tab 439 \tab 193 \tab 1468 \cr
   1983 \tab - \tab 295 \tab 525 \tab 531 \tab 248 \tab 1599 \cr
   1984 \tab - \tab 273 \tab 545 \tab 460 \tab 195 \tab 1473 \cr
   Total \tab 0 \tab 1130 \tab 2197 \tab 1913 \tab 832 \tab 6072 \cr
}
}

\seealso{
Other Core: 
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{sibdeath}},
\code{\link{spdeath}},
\code{\link{trarel1}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}
}
\concept{Core}
\keyword{variable}
