% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_84.R
\name{tratot1}
\alias{tratot1}
\title{Trauma last year, self & relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
tratot1
}
\details{
Question A. Total number of traumatic events happening to respondent and relatives last year.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/757/vshow}{GSS Data Explorer page for \code{tratot1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab no answer \tab 6 \tab Total \cr
   1978 \tab - \tab 409 \tab 493 \tab 363 \tab 167 \tab 33 \tab 11 \tab 56 \tab - \tab 1532 \cr
   1980 \tab - \tab 400 \tab 471 \tab 352 \tab 144 \tab 37 \tab 3 \tab 60 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 370 \tab 514 \tab 399 \tab 221 \tab 50 \tab 11 \tab 33 \tab 1 \tab 1599 \cr
   1984 \tab - \tab 358 \tab 500 \tab 344 \tab 148 \tab 47 \tab 8 \tab 68 \tab - \tab 1473 \cr
   Total \tab 0 \tab 1537 \tab 1978 \tab 1458 \tab 680 \tab 167 \tab 33 \tab 217 \tab 2 \tab 6072 \cr
}
}

\seealso{
Other Core: 
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{sibdeath}},
\code{\link{spdeath}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}
}
\concept{Core}
\keyword{variable}
