% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_265.R
\name{travel30}
\alias{travel30}
\title{R use travel site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
travel30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
H. Travel?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} 1-2 times
\item \verb{[3]} 3-5 times
\item \verb{[4]} more than 5 times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2401/vshow}{GSS Data Explorer page for \code{travel30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 230 \tab 233 \tab 111 \tab 95 \tab 73 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 366 \tab 298 \tab 171 \tab 154 \tab 94 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 231 \tab 218 \tab 121 \tab 108 \tab 12 \tab - \tab 2812 \cr
   Total \tab 5878 \tab 827 \tab 749 \tab 403 \tab 357 \tab 179 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{art30}},
\code{\link{games30}},
\code{\link{health30}},
\code{\link{humor30}},
\code{\link{music30}},
\code{\link{pol30}},
\code{\link{relig30}},
\code{\link{sports30}},
\code{\link{tvmov30}}
}
\concept{Information Society}
\keyword{variable}
