% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_278.R
\name{travelgn}
\alias{travelgn}
\title{R made travel reservations past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
travelgn
}
\details{
Question In the last twelve months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/2002"have you:
C. Made an airplane or other travel reservations?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2534/vshow}{GSS Data Explorer page for \code{travelgn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 1884 \tab 564 \tab 228 \tab 123 \tab 18 \tab - \tab - \tab 2817 \cr
   2002 \tab 1372 \tab 795 \tab 370 \tab 147 \tab 5 \tab 69 \tab 7 \tab 2765 \cr
   Total \tab 3256 \tab 1359 \tab 598 \tab 270 \tab 23 \tab 69 \tab 7 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinfgn}},
\code{\link{gamegn}},
\code{\link{investgn}},
\code{\link{newjobgn}},
\code{\link{peoplegn}},
\code{\link{polcamp}},
\code{\link{poldisgn}},
\code{\link{polfrel}},
\code{\link{polwww}}
}
\concept{Information Society}
\keyword{variable}
