% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{trcong}
\alias{trcong}
\title{Trust in congress, 10 point scale}
\source{
General Social Survey https://gss.norc.org
}
\description{
trcong
}
\details{
Question On a scale of 0 to 10, how much do you personally trust each of the following institutions? 0 means you do not trust an institution at all, and 10 means you trust it completely….The U.S. Congress.
}
\section{Values}{

\itemize{
\item \verb{[0]} no trust at all
\item \verb{[10]} complete trust
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8176/vshow}{GSS Data Explorer page for \code{trcong}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab not available in this year \tab no trust at all \tab 1 \tab complete trust \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab iap \tab no answer \tab Total \cr
   2024 \tab - \tab 262 \tab 83 \tab 17 \tab 156 \tab 236 \tab 152 \tab 338 \tab 127 \tab 69 \tab 45 \tab 6 \tab 127 \tab 1673 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 262 \tab 83 \tab 17 \tab 156 \tab 236 \tab 152 \tab 338 \tab 127 \tab 69 \tab 45 \tab 6 \tab 127 \tab 1673 \tab 18 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{dataprot}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{modeorg}},
\code{\link{trppl}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
