% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{trdestck}
\alias{trdestck}
\title{Company stock publicly traded}
\source{
General Social Survey https://gss.norc.org
}
\description{
trdestck
}
\details{
Question C. If your company has stock, is the stock of the company where your work traded on a stock market such as the New York Stock Exchange or the NASDAQ
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3563/vshow}{GSS Data Explorer page for \code{trdestck}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1765 \tab 123 \tab 530 \tab 115 \tab 5 \tab 2538 \cr
   Total \tab 1765 \tab 123 \tab 530 \tab 115 \tab 5 \tab 2538 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{defpensn}}
}
\concept{Shared Capitalism}
\keyword{variable}
