% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_634.R
\name{trdunio1}
\alias{trdunio1}
\title{Workers need strong trade unions to protect their interests}
\source{
General Social Survey https://gss.norc.org
}
\description{
trdunio1
}
\details{
Question Workers need strong trade unions to protect their interests
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7547/vshow}{GSS Data Explorer page for \code{trdunio1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2242 \tab 246 \tab 555 \tab 542 \tab 227 \tab 127 \tab 72 \tab 2 \tab 19 \tab 4032 \cr
   Total \tab 2242 \tab 246 \tab 555 \tab 542 \tab 227 \tab 127 \tab 72 \tab 2 \tab 19 \tab 4032 \cr
}
}

\keyword{variable}
