% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_297.R
\name{treat10}
\alias{treat10}
\title{R seek treat downhearted and blue}
\source{
General Social Survey https://gss.norc.org
}
\description{
treat10
}
\details{
Question Next, I’m going to read you some descriptions of health problems people sometimes experience. Some people might go to a doctor, others might not.
J. Feeling downhearted and blue?
}
\section{Values}{

\itemize{
\item \code{1} definitely go
\item \code{2} probably go
\item \code{3} probably not go
\item \code{4} definitely not go
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2697/vshow}{GSS Data Explorer page for \code{treat10}}.
}

\seealso{
Other Health Status: 
\code{\link{treat11}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{wait1}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}}

Other Health: 
\code{\link{treat11}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{wait1}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}}
}
\concept{Health}
\concept{Health Status}
\keyword{variable}
