% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{treat4}
\alias{treat4}
\title{R seek treat limited types of work}
\source{
General Social Survey https://gss.norc.org
}
\description{
treat4
}
\details{
Question Next, I'm going to read you some descriptions of health problems people sometimes experience. Some people might go to a doctor, others might not.
D. Being limited in the kind of work or other regular daily activities you Do as a result of your physical health?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely go
\item \verb{[2]} probably go
\item \verb{[3]} probably not go
\item \verb{[4]} definitely not go
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2691/vshow}{GSS Data Explorer page for \code{treat4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely go \tab probably go \tab probably not go \tab definitely not go \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 501 \tab 425 \tab 187 \tab 83 \tab 6 \tab 217 \tab 2817 \cr
   Total \tab 1398 \tab 501 \tab 425 \tab 187 \tab 83 \tab 6 \tab 217 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{pain}},
\code{\link{peaceful}},
\code{\link{socacts}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat5}}
}
\concept{Health Status}
\keyword{variable}
