% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_297.R
\name{treat8}
\alias{treat8}
\title{R seek treat not calm and peaceful}
\source{
General Social Survey https://gss.norc.org
}
\description{
treat8
}
\details{
Question Next, I'm going to read you some descriptions of health problems people sometimes experience. Some people might go to a doctor, others might not.
H. Not feeling calm and peaceful?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely go
\item \verb{[2]} probably go
\item \verb{[3]} probably not go
\item \verb{[4]} definitely not go
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2695/vshow}{GSS Data Explorer page for \code{treat8}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely go \tab probably go \tab probably not go \tab definitely not go \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 213 \tab 223 \tab 371 \tab 333 \tab 12 \tab 267 \tab 2817 \cr
   Total \tab 1398 \tab 213 \tab 223 \tab 371 \tab 333 \tab 12 \tab 267 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat9}},
\code{\link{wait1}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}}
}
\concept{Health Status}
\keyword{variable}
