% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{treatres}
\alias{treatres}
\title{People are treated with respect}
\source{
General Social Survey https://gss.norc.org
}
\description{
treatres
}
\details{
Question M. At work, people are treated with respect.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3010/vshow}{GSS Data Explorer page for \code{treatres}}.
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}}

Other Job: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}}
}
\concept{Job}
\concept{Workplace Conflict}
\keyword{variable}
