% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{trlegis}
\alias{trlegis}
\title{How much do you trust the u.s. congress?}
\source{
General Social Survey https://gss.norc.org
}
\description{
trlegis
}
\details{
Question How much do you trust the U.S. Congress?
}
\section{Values}{

\itemize{
\item \verb{[0]} 0 - no trust
\item \verb{[10]} 10 - complete trust
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7485/vshow}{GSS Data Explorer page for \code{trlegis}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 0 - no trust \tab 1 \tab 10 - complete trust \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 339 \tab 195 \tab 25 \tab 197 \tab 255 \tab 222 \tab 332 \tab 108 \tab 88 \tab 46 \tab 11 \tab 49 \tab 156 \tab 6 \tab 4032 \cr
   Total \tab 2003 \tab 339 \tab 195 \tab 25 \tab 197 \tab 255 \tab 222 \tab 332 \tab 108 \tab 88 \tab 46 \tab 11 \tab 49 \tab 156 \tab 6 \tab 4032 \cr
}
}

\keyword{variable}
