% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{trmedia}
\alias{trmedia}
\title{How much do you trust the news media?}
\source{
General Social Survey https://gss.norc.org
}
\description{
trmedia
}
\details{
Question How much do you trust the news media?
}
\section{Values}{

\itemize{
\item \code{0} 0 - no trust
\item \code{10} 10 - complete trust
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7483/vshow}{GSS Data Explorer page for \code{trmedia}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 0 - no trust \tab 1 \tab 10 - complete trust \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 372 \tab 149 \tab 19 \tab 199 \tab 179 \tab 150 \tab 291 \tab 153 \tab 172 \tab 105 \tab 31 \tab 45 \tab 155 \tab 9 \tab 4032 \cr
   Total \tab 2003 \tab 372 \tab 149 \tab 19 \tab 199 \tab 179 \tab 150 \tab 291 \tab 153 \tab 172 \tab 105 \tab 31 \tab 45 \tab 155 \tab 9 \tab 4032 \cr
}
}

\keyword{variable}
