% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_404.R
\name{trtasian}
\alias{trtasian}
\title{Number of asian r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtasian
}
\details{
Question Now I’m going to ask you some questions about people that you trust, for example good friends, people you discuss important matters with, or trust for advice, or trust with money. Some of these questions may seem unusual but they are an important way to help us understand more about social networks in America. Please answer the questions as best you can.
U. How many are Asian or Asian-American?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3686/vshow}{GSS Data Explorer page for \code{trtasian}}.
}

\seealso{
Other Number Known: 
\code{\link{trtblack}},
\code{\link{trtchild}},
\code{\link{trtelecs}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtmils}},
\code{\link{trtprisn}},
\code{\link{trtunemp}}

Other Social Networks: 
\code{\link{trtblack}},
\code{\link{trtchild}},
\code{\link{trtelecs}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtjans}},
\code{\link{trtmils}},
\code{\link{trtprisn}},
\code{\link{trtunemp}}
}
\concept{Asian Americans}
\concept{Number Known}
\concept{Social Networks}
\keyword{variable}
