% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_405.R
\name{trtdems}
\alias{trtdems}
\title{Proportion of democrat r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtdems
}
\details{
Question B. How many do you think are Democrats? (Would you say…)
}
\section{Values}{

\itemize{
\item \verb{[1]} almost all
\item \verb{[2]} most
\item \verb{[3]} about half
\item \verb{[4]} a few
\item \verb{[5]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3697/vshow}{GSS Data Explorer page for \code{trtdems}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost all \tab most \tab about half \tab a few \tab none \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 59 \tab 76 \tab 208 \tab 176 \tab 40 \tab 114 \tab 16 \tab 4510 \cr
   Total \tab 3821 \tab 59 \tab 76 \tab 208 \tab 176 \tab 40 \tab 114 \tab 16 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{trtattnd}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trtlib}},
\code{\link{trtnoatt}},
\code{\link{trtreps}},
\code{\link{trtwhite}}
}
\concept{Number Known}
\keyword{variable}
