% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_404.R
\name{trtjans}
\alias{trtjans}
\title{Number of janitor r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtjans
}
\details{
Question Now I’m going to ask you some questions about people that you trust, for example good friends, people you discuss important matters with, or trust for advice, or trust with money. Some of these questions may seem unusual but they are an important way to help us understand more about social networks in America. Please answer the questions as best you can.
N. How many are janitors or building cleaners?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3679/vshow}{GSS Data Explorer page for \code{trtjans}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 418 \tab 117 \tab 106 \tab 12 \tab 2 \tab 20 \tab 14 \tab 4510 \cr
   Total \tab 3821 \tab 418 \tab 117 \tab 106 \tab 12 \tab 2 \tab 20 \tab 14 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{trtasian}},
\code{\link{trtblack}},
\code{\link{trtchild}},
\code{\link{trtelecs}},
\code{\link{trthisp}},
\code{\link{trthome}},
\code{\link{trtmils}},
\code{\link{trtprisn}},
\code{\link{trtunemp}}
}
\concept{Number Known}
\keyword{variable}
