% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_403.R
\name{trtkeith}
\alias{trtkeith}
\title{Number ofkeith r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtkeith
}
\details{
Question Now I'm going to ask you some questions about people that you trust, for example good friends, people you discuss important matters with, or trust for advice, or trust with money. Some of these questions may seem unusual but they are an important way to help us understand more about social networks in America. Please answer the questions as best you can.
E. How many are named Keith?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3670/vshow}{GSS Data Explorer page for \code{trtkeith}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 469 \tab 141 \tab 43 \tab 3 \tab 18 \tab 15 \tab 4510 \cr
   Total \tab 3821 \tab 469 \tab 141 \tab 43 \tab 3 \tab 18 \tab 15 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{trtbrnda}},
\code{\link{trtcops}},
\code{\link{trtjose}},
\code{\link{trtlaws}},
\code{\link{trtlinda}},
\code{\link{trtmaria}},
\code{\link{trtmark}},
\code{\link{trtrachl}},
\code{\link{trtsocs}}
}
\concept{Number Known}
\keyword{variable}
