% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{trtmyrac}
\alias{trtmyrac}
\title{Proportion of same race as r among ppl r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtmyrac
}
\details{
Question Are the people that you trust…
}
\section{Values}{

\itemize{
\item \verb{[1]} almost all the same race as you
\item \verb{[2]} mostly the same race as you
\item \verb{[3]} about evenly divided
\item \verb{[4]} mostly a different race than you
\item \verb{[5]} almost all a different race than you
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3699/vshow}{GSS Data Explorer page for \code{trtmyrac}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost all the same race as you \tab mostly the same race as you \tab about evenly divided \tab mostly a different race than you \tab almost all a different race than you \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 267 \tab 218 \tab 140 \tab 16 \tab 11 \tab 24 \tab 13 \tab 4510 \cr
   Total \tab 3821 \tab 267 \tab 218 \tab 140 \tab 16 \tab 11 \tab 24 \tab 13 \tab 4510 \cr
}
}

\concept{Number Known}
\keyword{variable}
