% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_405.R
\name{trtnoatt}
\alias{trtnoatt}
\title{Number of non church attendeer trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtnoatt
}
\details{
Question Now I'm going to ask you some questions about people that you trust, for example good friends, people you discuss important matters with, or trust for advice, or trust with money. Some of these questions may seem unusual but they are an important way to help us understand more about social networks in America. Please answer the questions as best you can.
AB. How many people are you pretty certain attend religious services rarely or never?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1
\item \verb{[3]} 2 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} more than 10
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3693/vshow}{GSS Data Explorer page for \code{trtnoatt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 100 \tab 59 \tab 225 \tab 71 \tab 136 \tab 82 \tab 16 \tab 4510 \cr
   Total \tab 3821 \tab 100 \tab 59 \tab 225 \tab 71 \tab 136 \tab 82 \tab 16 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{trtattnd}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtdems}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trtlib}},
\code{\link{trtreps}},
\code{\link{trtwhite}}
}
\concept{Number Known}
\keyword{variable}
