% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_405.R
\name{trtreps}
\alias{trtreps}
\title{Proportion of republican r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtreps
}
\details{
Question A. How many of the people that you trust do you think are Republicans? Would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} almost all
\item \verb{[2]} most
\item \verb{[3]} about half
\item \verb{[4]} a few
\item \verb{[5]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3696/vshow}{GSS Data Explorer page for \code{trtreps}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost all \tab most \tab about half \tab a few \tab none \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 50 \tab 81 \tab 189 \tab 151 \tab 90 \tab 111 \tab 17 \tab 4510 \cr
   Total \tab 3821 \tab 50 \tab 81 \tab 189 \tab 151 \tab 90 \tab 111 \tab 17 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{trtattnd}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtdems}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trtlib}},
\code{\link{trtnoatt}},
\code{\link{trtwhite}}
}
\concept{Number Known}
\keyword{variable}
