% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{trust1}
\alias{trust1}
\title{Can people be trusted (see codebook)}
\source{
General Social Survey https://gss.norc.org
}
\description{
trust1
}
\details{
Question C. Generally speaking, would you say that most people can be trusted or that you need to be very careful in dealing with people?
}
\section{Values}{

\itemize{
\item \code{1} can trust
\item \code{2} need to be very careful
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6644/vshow}{GSS Data Explorer page for \code{trust1}}.
}

\keyword{variable}
