% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{trustcrt}
\alias{trustcrt}
\title{Would trust the courts}
\source{
General Social Survey https://gss.norc.org
}
\description{
trustcrt
}
\details{
Question People sometimes are incapable of making decisions about their care and medical treatment at the end of life. If you were incapable, how much trust would you put in the following people to do what was best for you?
READ RESPONSE CATEGORIES
C. I would trust the courts.
}
\section{Values}{

\itemize{
\item \verb{[1]} completely
\item \verb{[2]} a great deal
\item \verb{[3]} somewhat
\item \verb{[4]} only a little
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2253/vshow}{GSS Data Explorer page for \code{trustcrt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab completely \tab a great deal \tab somewhat \tab only a little \tab not at all \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 51 \tab 55 \tab 372 \tab 284 \tab 579 \tab 42 \tab 4 \tab 2832 \cr
   Total \tab 1445 \tab 51 \tab 55 \tab 372 \tab 284 \tab 579 \tab 42 \tab 4 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{painmed}},
\code{\link{rightref}},
\code{\link{termvig}},
\code{\link{trustdoc}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\keyword{variable}
