% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{trustman}
\alias{trustman}
\title{R trust management at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
trustman
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
J. I trust the management at the place where I work
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2783/vshow}{GSS Data Explorer page for \code{trustman}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 548 \tab 818 \tab 298 \tab 101 \tab 13 \tab 18 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 534 \tab 785 \tab 269 \tab 100 \tab 28 \tab 18 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 350 \tab 595 \tab 157 \tab 55 \tab 3 \tab 27 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 369 \tab 656 \tab 172 \tab 35 \tab 9 \tab 9 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 355 \tab 784 \tab 215 \tab 51 \tab 12 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 556 \tab 1016 \tab 274 \tab 95 \tab 19 \tab 15 \tab 23 \tab 3544 \cr
   Total \tab 8336 \tab 2712 \tab 4654 \tab 1385 \tab 437 \tab 84 \tab 118 \tab 23 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
