% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{trustpeo}
\alias{trustpeo}
\title{There are only a few ppl r can trust completely}
\source{
General Social Survey https://gss.norc.org
}
\description{
trustpeo
}
\details{
Question To what extent do you agree or disagree with the following statements?
A. There are only a few people I can trust completely
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4236/vshow}{GSS Data Explorer page for \code{trustpeo}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{demands}},
\code{\link{helpfrds}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
