% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{trustv}
\alias{trustv}
\title{Can people be trusted (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
trustv
}
\details{
Question Generally speaking, would you say that most people can be trusted or that you can't
be too careful in dealing with people?
}
\section{Values}{

\itemize{
\item \verb{[1]} most people can be trusted
\item \verb{[2]} can't be too careful
\item \verb{[3]} depends (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7593/vshow}{GSS Data Explorer page for \code{trustv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab most people can be trusted \tab can't be too careful \tab depends (phone mode only: volunteered) \tab skipped on web \tab Total \cr
   2021 \tab 2673 \tab 295 \tab 557 \tab 505 \tab 2 \tab 4032 \cr
   2022 \tab 2974 \tab 98 \tab 257 \tab 214 \tab 1 \tab 3544 \cr
   2024 \tab 2716 \tab 96 \tab 275 \tab 216 \tab 6 \tab 3309 \cr
   Total \tab 8363 \tab 489 \tab 1089 \tab 935 \tab 9 \tab 10885 \cr
}
}

\keyword{variable}
