% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{trusty}
\alias{trusty}
\title{Can people be trusted-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
trusty
}
\details{
Question B. Do you think most people can be trusted?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/442/vshow}{GSS Data Explorer page for \code{trusty}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1983 \tab 804 \tab 443 \tab 326 \tab 21 \tab 5 \tab 1599 \cr
   Total \tab 804 \tab 443 \tab 326 \tab 21 \tab 5 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{fair}},
\code{\link{hapcohab}},
\code{\link{health}},
\code{\link{helpful}},
\code{\link{life}},
\code{\link{satcity}},
\code{\link{sathobby}},
\code{\link{trust}}
}
\concept{Core}
\keyword{variable}
