% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{trynewjb}
\alias{trynewjb}
\title{How likely r make effort for new job next year}
\source{
General Social Survey https://gss.norc.org
}
\description{
trynewjb
}
\details{
Question Taking everything into consideration, how likely is it you will make a genuine effort to find a new job with another employer within the next year?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2006, 2010, 2012, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2820/vshow}{GSS Data Explorer page for \code{trynewjb}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not at all likely \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 330 \tab 355 \tab 1085 \tab 12 \tab 14 \tab - \tab 2765 \cr
   2004 \tab 945 \tab 301 \tab 375 \tab 1152 \tab 4 \tab 35 \tab - \tab 2812 \cr
   2006 \tab 2776 \tab 283 \tab 362 \tab 1058 \tab 18 \tab 13 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 222 \tab 256 \tab 680 \tab 5 \tab 24 \tab - \tab 2044 \cr
   2012 \tab 795 \tab 214 \tab 215 \tab 713 \tab 7 \tab 30 \tab - \tab 1974 \cr
   2014 \tab 1288 \tab 214 \tab 254 \tab 776 \tab 1 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 246 \tab 353 \tab 814 \tab 3 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 300 \tab 446 \tab 1192 \tab 12 \tab 18 \tab 21 \tab 3544 \cr
   Total \tab 10085 \tab 2110 \tab 2616 \tab 7470 \tab 62 \tab 171 \tab 21 \tab 22535 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}},
\code{\link{wksexism}}

Other Workplace Conflict: 
\code{\link{health1}},
\code{\link{wkageism}},
\code{\link{wkracism}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
