% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{tvhours}
\alias{tvhours}
\title{Hours per day watching tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvhours
}
\details{
Question On the average day, about how many hours do you personally watch television?
}
\section{Values}{

\itemize{
\item \code{0} 0 hours
\item \code{24} 24 hours
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1988 \tab A/B/- \tab partial \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/691/vshow}{GSS Data Explorer page for \code{tvhours}}.
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}}
}
\concept{Core}
\concept{Television Viewing}
\keyword{variable}
