% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_378.R
\name{tvinuit}
\alias{tvinuit}
\title{R would choose tv program about native ppl in the arctic}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvinuit
}
\details{
Question The next few questions are about the regions near the poles. Please look at Card B22. If you were choosing a television program to watch, would you be very likely, somewhat likely, somewhat unlikely, or very unlikely to watch a program about…
B. Native peoples living in the Arctic?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3420/vshow}{GSS Data Explorer page for \code{tvinuit}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{tvbears}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}}

Other Science: 
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{tvbears}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}}

Other Television Viewing: 
\code{\link{tvbears}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}}
}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\concept{Television Viewing}
\keyword{variable}
