% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{tvnews1}
\alias{tvnews1}
\title{Reliability of news on tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvnews1
}
\details{
Question In general, how reliable would you rate the news on political matters found in the following sources?… Television (including TV apps or access via Internet).
}
\section{Values}{

\itemize{
\item \code{1} very unreliable
\item \code{2} somewhat unreliable
\item \code{3} somewhat reliable
\item \code{4} very reliable
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8166/vshow}{GSS Data Explorer page for \code{tvnews1}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modeorg}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{radionews}},
\code{\link{trppl}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
