% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_378.R
\name{tvoil}
\alias{tvoil}
\title{R would choose tv program about oil drilling in the arctic}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvoil
}
\details{
Question The next few questions are about the regions near the poles. Please look at Card B22. If you were choosing a television program to watch, would you be very likely, somewhat likely, somewhat unlikely, or very unlikely to watch a program about…
E. Oil drilling in the Arctic?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3423/vshow}{GSS Data Explorer page for \code{tvoil}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 481 \tab 645 \tab 361 \tab 351 \tab 24 \tab 2 \tab 4510 \cr
   Total \tab 2646 \tab 481 \tab 645 \tab 361 \tab 351 \tab 24 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvozone}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
