% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{tvpbs}
\alias{tvpbs}
\title{How often r watches public tv shows}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvpbs
}
\details{
Question I'll now ask about some different kinds of television shows. Would you tell me how often you watch prime-time drama or situation comedy programs?  Would you say every day, several times a week, several times a month, rarely, or never?
C. Programs shown on public television.
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} sevrl days in week
\item \verb{[3]} sevrl days in month
\item \verb{[4]} rarely
\item \verb{[5]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1421/vshow}{GSS Data Explorer page for \code{tvpbs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab daily \tab sevrl days in week \tab sevrl days in month \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 298 \tab 477 \tab 351 \tab 307 \tab 158 \tab 7 \tab 8 \tab 1606 \cr
   Total \tab 0 \tab 298 \tab 477 \tab 351 \tab 307 \tab 158 \tab 7 \tab 8 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{judgeart}},
\code{\link{modpaint}},
\code{\link{pclit}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvshows}}
}
\concept{Culture}
\keyword{variable}
