% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_116.R
\name{tvrelig}
\alias{tvrelig}
\title{Hours per week r watches religious tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvrelig
}
\details{
Question About how much time per week, in hours and minutes, do you normally spend watching religious shows on television?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1075/vshow}{GSS Data Explorer page for \code{tvrelig}}.
}

\seealso{
Other Religion: 
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{god}},
\code{\link{grace16}},
\code{\link{join16}},
\code{\link{lapsed}},
\code{\link{reborn}},
\code{\link{savesoul}},
\code{\link{saygrace}}
}
\concept{Religion}
\keyword{variable}
