% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{twoincs}
\alias{twoincs}
\title{Husb & wife should contrib to family income}
\source{
General Social Survey https://gss.norc.org
}
\description{
twoincs
}
\details{
Question Do you agree or disagree ...
H. Both the husband and the wife should contribute to the household income.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4350/vshow}{GSS Data Explorer page for \code{twoincs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 210 \tab 463 \tab 414 \tab 223 \tab 44 \tab 31 \tab 29 \tab 1481 \cr
   1994 \tab 1545 \tab 277 \tab 527 \tab 417 \tab 157 \tab 17 \tab 26 \tab 26 \tab 2992 \cr
   1996 \tab 1444 \tab 440 \tab 536 \tab 303 \tab 138 \tab 24 \tab 17 \tab 2 \tab 2904 \cr
   Total \tab 3056 \tab 927 \tab 1526 \tab 1134 \tab 518 \tab 85 \tab 74 \tab 57 \tab 7377 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famsuffr}},
\code{\link{fejobind}},
\code{\link{hapifwrk}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{kidsuffr}},
\code{\link{mawrkwrm}}
}
\concept{Gender}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
