% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{twoincs1}
\alias{twoincs1}
\title{Both men and women should contribute to income}
\source{
General Social Survey https://gss.norc.org
}
\description{
twoincs1
}
\details{
Question A. Do you agree or disagree ... both the husband and the wife should contribute to the household income.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4455/vshow}{GSS Data Explorer page for \code{twoincs1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 277 \tab 527 \tab 417 \tab 157 \tab 17 \tab 26 \tab 26 \tab 2992 \cr
   2002 \tab 1594 \tab 341 \tab 327 \tab 369 \tab - \tab 109 \tab 23 \tab 2 \tab 2765 \cr
   2012 \tab 672 \tab 249 \tab 572 \tab 307 \tab 121 \tab 18 \tab 28 \tab 7 \tab 1974 \cr
   Total \tab 3811 \tab 867 \tab 1426 \tab 1093 \tab 278 \tab 144 \tab 77 \tab 35 \tab 7731 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
