% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{twokids}
\alias{twokids}
\title{Desirability of two-kid family}
\source{
General Social Survey https://gss.norc.org
}
\description{
twokids
}
\details{
Question In general, what do you feel about each of these family sizes?
C. A family with two children
}
\section{Values}{

\itemize{
\item \verb{[1]} very desirable
\item \verb{[2]} desirable
\item \verb{[3]} neither
\item \verb{[4]} undesirable
\item \verb{[5]} very undesirable
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4377/vshow}{GSS Data Explorer page for \code{twokids}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very desirable \tab desirable \tab neither \tab undesirable \tab very undesirable \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 358 \tab 726 \tab 169 \tab 32 \tab 13 \tab 48 \tab 68 \tab 1481 \cr
   Total \tab 67 \tab 358 \tab 726 \tab 169 \tab 32 \tab 13 \tab 48 \tab 68 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{mardiv}},
\code{\link{marhomo}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{threkids}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
