% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_374.R
\name{txpol}
\alias{txpol}
\title{How well elected official knows effect of reducing fed inc tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
txpol
}
\details{
Question The first issue is federal income taxes. Some people say that reducing income taxes almost always helps the economy grow. Others say that reducing income taxes has very little long-term impact on economic growth but could cause a big increase in the deficit. Please look at Card B10. (On a scale of 1 to 5, where  1 means "Very Well" and 5 means "Not at All"), how well do the following groups understand the likely effects of reducing federal income taxes?
C. Elected Officials
}
\section{Values}{

\itemize{
\item \verb{[1]} very well
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3379/vshow}{GSS Data Explorer page for \code{txpol}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very well \tab 2 \tab 3 \tab 4 \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3574 \tab 106 \tab 143 \tab 320 \tab 177 \tab 143 \tab 45 \tab 2 \tab 4510 \cr
   2010 \tab 1578 \tab 52 \tab 67 \tab 155 \tab 93 \tab 73 \tab 23 \tab 3 \tab 2044 \cr
   Total \tab 5152 \tab 158 \tab 210 \tab 475 \tab 270 \tab 216 \tab 68 \tab 5 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbsttx}},
\code{\link{bizinftx}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{polbsttx}},
\code{\link{polinftx}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
