% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{typdealr}
\alias{typdealr}
\title{Rwas the agency a new or used car dealer'ship?}
\source{
General Social Survey https://gss.norc.org
}
\description{
typdealr
}
\details{
Question Was the agency a new car dealership or did it only sell used cars?
}
\section{Values}{

\itemize{
\item \verb{[1]} new car dealership
\item \verb{[2]} used car dealership
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1801/vshow}{GSS Data Explorer page for \code{typdealr}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab new car dealership \tab used car dealership \tab no answer \tab Total \cr
   1996 \tab 2706 \tab 114 \tab 72 \tab 12 \tab 2904 \cr
   Total \tab 2706 \tab 114 \tab 72 \tab 12 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{newused}},
\code{\link{relprivt}},
\code{\link{warrntyc}}
}
\concept{Markets}
\keyword{variable}
