% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{typeurl}
\alias{typeurl}
\title{How often r types urls by self}
\source{
General Social Survey https://gss.norc.org
}
\description{
typeurl
}
\details{
Question Now I’d like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit:
B. Type (or paste in) the web address yourself?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than half of the time
\item \code{3} about half of the time
\item \code{4} more than half of the time
\item \code{5} almost always
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2389/vshow}{GSS Data Explorer page for \code{typeurl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than half of the time \tab about half of the time \tab more than half of the time \tab almost always \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 90 \tab 188 \tab 135 \tab 113 \tab 135 \tab 81 \tab - \tab 2817 \cr
   2002 \tab 2403 \tab 30 \tab 91 \tab 90 \tab 40 \tab 61 \tab 49 \tab 1 \tab 2765 \cr
   Total \tab 4478 \tab 120 \tab 279 \tab 225 \tab 153 \tab 196 \tab 130 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{drctlink}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{srcheng}},
\code{\link{usewww}},
\code{\link{wwwhr}},
\code{\link{wwwmin}}
}
\concept{Information Society}
\keyword{variable}
