% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{typevisa}
\alias{typevisa}
\title{What kind of visa}
\source{
General Social Survey https://gss.norc.org
}
\description{
typevisa
}
\details{
Question What kind of visa?
}
\section{Values}{

\itemize{
\item \verb{[1]} h1b
\item \verb{[2]} l1
\item \verb{[3]} student
\item \verb{[4]} tourist
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3901/vshow}{GSS Data Explorer page for \code{typevisa}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab h1b \tab l1 \tab student \tab tourist \tab other \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2008 \tab 1982 \tab 6 \tab 2 \tab 4 \tab 2 \tab 22 \tab 4 \tab 1 \tab - \tab 2023 \cr
   Total \tab 1982 \tab 6 \tab 2 \tab 4 \tab 2 \tab 22 \tab 4 \tab 1 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{autonojb}},
\code{\link{bornsp}},
\code{\link{curempyr}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{natdeg}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work3yrs}}

Other Citizenship: 
\code{\link{actassoc}},
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{buypol}},
\code{\link{fucitzn}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{notvote}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{obeylaws}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obmepax}},
\code{\link{obmewar}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}},
\code{\link{visa}},
\code{\link{voteelec}},
\code{\link{watchgov}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornhome}},
\code{\link{bornsp}},
\code{\link{fucitzn}},
\code{\link{immassim}},
\code{\link{immcrmup}},
\code{\link{immcult}},
\code{\link{immecon}},
\code{\link{immeduc}},
\code{\link{immfare}},
\code{\link{immimp}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{immpush}},
\code{\link{immrghts}},
\code{\link{immstats}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{letin}},
\code{\link{letin1a}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{natborn}},
\code{\link{natdeg}},
\code{\link{patriot4}},
\code{\link{rimmdisc}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{visa}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Citizenship}
\concept{Global Economics}
\concept{Immigration}
\keyword{variable}
