% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{typfrat}
\alias{typfrat}
\title{R belongs to more than one fraternal groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
typfrat
}
\details{
Question A. Do you belong to more than one fraternal group?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3110/vshow}{GSS Data Explorer page for \code{typfrat}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{churchtx}},
\code{\link{infrmgrp}},
\code{\link{typpolit}},
\code{\link{typserv}},
\code{\link{typvet}}

Other Voluntary Associations: 
\code{\link{churchtx}},
\code{\link{infrmgrp}},
\code{\link{typpolit}},
\code{\link{typserv}},
\code{\link{typvet}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
