% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_347.R
\name{typnat}
\alias{typnat}
\title{R belongs to more than one nationality groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
typnat
}
\details{
Question K. Do you belong to more than one nationality club?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3120/vshow}{GSS Data Explorer page for \code{typnat}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2768 \tab 2 \tab 36 \tab 6 \tab 2812 \cr
   Total \tab 2768 \tab 2 \tab 36 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{typfarm}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typlit}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typsport}},
\code{\link{typunion}},
\code{\link{typyouth}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
