% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_347.R
\name{typprof}
\alias{typprof}
\title{R belongs to more than one professional societies}
\source{
General Social Survey https://gss.norc.org
}
\description{
typprof
}
\details{
Question N. Do you belong to more than one professional or academic society?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3123/vshow}{GSS Data Explorer page for \code{typprof}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2586 \tab 73 \tab 143 \tab 10 \tab 2812 \cr
   Total \tab 2586 \tab 73 \tab 143 \tab 10 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{typfarm}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typschl}},
\code{\link{typsport}},
\code{\link{typunion}},
\code{\link{typyouth}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
