% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{typserv}
\alias{typserv}
\title{R belongs to more than one service clubs}
\source{
General Social Survey https://gss.norc.org
}
\description{
typserv
}
\details{
Question B. Do you belong to more than one service club?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3111/vshow}{GSS Data Explorer page for \code{typserv}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2666 \tab 32 \tab 108 \tab 6 \tab 2812 \cr
   Total \tab 2666 \tab 32 \tab 108 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{churchtx}},
\code{\link{infrmgrp}},
\code{\link{typfrat}},
\code{\link{typpolit}},
\code{\link{typvet}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
