% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_347.R
\name{typunion}
\alias{typunion}
\title{R belongs to more than one labor unions}
\source{
General Social Survey https://gss.norc.org
}
\description{
typunion
}
\details{
Question E. Do you belong to more than one labor union?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3114/vshow}{GSS Data Explorer page for \code{typunion}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{typfarm}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typsport}},
\code{\link{typyouth}}

Other Voluntary Associations: 
\code{\link{typfarm}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typsport}},
\code{\link{typyouth}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
