% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{unaunum}
\alias{unaunum}
\title{Number of aunts and uncles}
\source{
General Social Survey https://gss.norc.org
}
\description{
unaunum
}
\details{
Question Now thinking of all your adult relatives -- those still living and age 18 or older.
How many of each do you have?  READ EACH ITEM. CODE A NUMBER OR ZERO FOR EACH.
C. Aunts, Uncles
}
\section{Values}{

\itemize{
\item \code{96} 96 or more
\item \code{97} # dk but ) 0
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4155/vshow}{GSS Data Explorer page for \code{unaunum}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 24 \tab 25 \tab 26 \tab 28 \tab 3 \tab 30 \tab 32 \tab 34 \tab 35 \tab 36 \tab 38 \tab 4 \tab 40 \tab 48 \tab 5 \tab 50 \tab 55 \tab 6 \tab 64 \tab 7 \tab 70 \tab 8 \tab 9 \tab 96 or more \tab # dk but ) 0 \tab don't know \tab no answer \tab Total \cr
   1986 \tab - \tab 298 \tab 124 \tab 76 \tab 16 \tab 46 \tab 15 \tab 13 \tab 33 \tab 19 \tab 15 \tab 12 \tab 2 \tab 117 \tab 53 \tab 2 \tab 5 \tab 7 \tab 9 \tab 2 \tab 4 \tab 95 \tab 8 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 122 \tab 5 \tab 1 \tab 111 \tab 4 \tab 1 \tab 86 \tab 1 \tab 51 \tab 1 \tab 57 \tab 33 \tab 1 \tab 1 \tab 9 \tab 8 \tab 1470 \cr
   Total \tab 0 \tab 298 \tab 124 \tab 76 \tab 16 \tab 46 \tab 15 \tab 13 \tab 33 \tab 19 \tab 15 \tab 12 \tab 2 \tab 117 \tab 53 \tab 2 \tab 5 \tab 7 \tab 9 \tab 2 \tab 4 \tab 95 \tab 8 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 122 \tab 5 \tab 1 \tab 111 \tab 4 \tab 1 \tab 86 \tab 1 \tab 51 \tab 1 \tab 57 \tab 33 \tab 1 \tab 1 \tab 9 \tab 8 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
