% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{unbetter}
\alias{unbetter}
\title{Working condition would be much worse without unions}
\source{
General Social Survey https://gss.norc.org
}
\description{
unbetter
}
\details{
Question To what extent do you agree or disagree with the following statements?
B. Without trade unions the working conditions of employees would be much worse than they are
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4609/vshow}{GSS Data Explorer page for \code{unbetter}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 261 \tab 606 \tab 321 \tab 226 \tab 61 \tab 41 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 261 \tab 606 \tab 321 \tab 226 \tab 61 \tab 41 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{econsup}},
\code{\link{rimpskls}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unjobsec}},
\code{\link{wantads}}
}
\concept{Work Orientation}
\keyword{variable}
