% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{unbored}
\alias{unbored}
\title{Rank of boredom when unemployed}
\source{
General Social Survey https://gss.norc.org
}
\description{
unbored
}
\details{
Question Suppose you were unemployed and couldn't find a job. Which of the following problems would be the worst?  Please write '1' in the box beside the worst thing, write '2' in the box beside the next worst thing, and '3' beside the third worst thing. Leave the other boxes blank.
G. Not knowing how to fill one's time
}
\section{Values}{

\itemize{
\item \verb{[1]} worst
\item \verb{[2]} 2nd worst
\item \verb{[3]} 3rd worst
\item \verb{[4]} not chosen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4522/vshow}{GSS Data Explorer page for \code{unbored}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab worst \tab 2nd worst \tab 3rd worst \tab not chosen \tab don't know \tab no answer \tab Total \cr
   1989 \tab 84 \tab 20 \tab 49 \tab 112 \tab 930 \tab 56 \tab 286 \tab 1537 \cr
   Total \tab 84 \tab 20 \tab 49 \tab 112 \tab 930 \tab 56 \tab 286 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
