% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{undk}
\alias{undk}
\title{Don't know response to unemployment items}
\source{
General Social Survey https://gss.norc.org
}
\description{
undk
}
\details{
Question Suppose you were unemployed and couldn't find a job. Which of the following problems would be the worst?  Please write '1' in the box beside the worst thing, write '2' in the box beside the next worst thing, and '3' beside the third worst thing. Leave the other boxes blank.
H. Respondent can't choose
}
\section{Values}{

\itemize{
\item \code{1} dk chosen
\item \code{2} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4523/vshow}{GSS Data Explorer page for \code{undk}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab dk chosen \tab not chosen \tab no answer \tab Total \cr
   1989 \tab 84 \tab 56 \tab 1198 \tab 199 \tab 1537 \cr
   Total \tab 84 \tab 56 \tab 1198 \tab 199 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
