% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_173.R
\name{undockid}
\alias{undockid}
\title{Us citizenship for children of illegal immigrants}
\source{
General Social Survey https://gss.norc.org
}
\description{
undockid
}
\details{
Question Under current law, immigrants who come from other countries to the United States legally are entitled, from the very beginning, to government assistance such as Medicaid, food stamps, or welfare on the same basis as citizens. But some people say they should not be eligible until they have lived here for a year or more. Which do you think? Do you think that immigrants who are here legally should be eligible for such services as soon as they come, or should they not be eligible?'
A. What about "undocumented aliens," that is, those who have immigrated to this country illegally?
3. Should they be entitled to have their children continue to qualify as American citizens if born in the United States, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes, qualify
\item \code{2} no, not qualify
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1619/vshow}{GSS Data Explorer page for \code{undockid}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{hspasn10}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{undoccol}},
\code{\link{undocwrk}}
}
\concept{Multiculturalism}
\keyword{variable}
