% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_173.R
\name{undocwrk}
\alias{undocwrk}
\title{Should illegal immigrants get work permits?}
\source{
General Social Survey https://gss.norc.org
}
\description{
undocwrk
}
\details{
Question Under current law, immigrants who come from other countries to the United States legally are entitled, from the very beginning, to government assistance such as Medicaid, food stamps, or welfare on the same basis as citizens. But some people say they should not be eligible until they have lived here for a year or more. Which do you think? Do you think that immigrants who are here legally should be eligible for such services as soon as they come, or should they not be eligible?'
A. What about "undocumented aliens," that is, those who have immigrated to this country illegally?
\enumerate{
\item Should illegal immigrants be entitled to work permits, or not?
}
}
\section{Values}{

\itemize{
\item \code{1} yes, entitled
\item \code{2} no, not entitled
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1617/vshow}{GSS Data Explorer page for \code{undocwrk}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{asngovt}},
\code{\link{bilinged}},
\code{\link{blkgovt}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{econfutr}},
\code{\link{econpast}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethhist}},
\code{\link{ethid}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{hspasn10}},
\code{\link{hspgovt}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{jobaff}},
\code{\link{letin}},
\code{\link{meltpot}},
\code{\link{obrespct}},
\code{\link{owneth}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}},
\code{\link{schleth}},
\code{\link{symptblk}},
\code{\link{teacheth}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{united}},
\code{\link{whoteach}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Multiculturalism}
\keyword{variable}
