% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{unhlthy}
\alias{unhlthy}
\title{How often does r work in unhealthy conds}
\source{
General Social Survey https://gss.norc.org
}
\description{
unhlthy
}
\details{
Question Now some more questions about your working conditions. one code for each item below to show how often it applies to your work. How often...
F. Do you work in unhealthy conditions?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} hardly ever
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4555/vshow}{GSS Data Explorer page for \code{unhlthy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 21 \tab 43 \tab 156 \tab 218 \tab 429 \tab 6 \tab 25 \tab 1537 \cr
   Total \tab 639 \tab 21 \tab 43 \tab 156 \tab 218 \tab 429 \tab 6 \tab 25 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
