% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{uniongod}
\alias{uniongod}
\title{Desire to be closer to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
uniongod
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
D. I desire to be closer to or in union with God.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never/almst nevr
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2111/vshow}{GSS Data Explorer page for \code{uniongod}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never/almst nevr \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 230 \tab 387 \tab 227 \tab 204 \tab 182 \tab 158 \tab 22 \tab 35 \tab 2832 \cr
   Total \tab 1387 \tab 230 \tab 387 \tab 227 \tab 204 \tab 182 \tab 158 \tab 22 \tab 35 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{beausprt}},
\code{\link{effort}},
\code{\link{fidgety}},
\code{\link{godlove}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
