% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{unionsbd}
\alias{unionsbd}
\title{Strong trade unions are bad for us economy}
\source{
General Social Survey https://gss.norc.org
}
\description{
unionsbd
}
\details{
Question (HAND CARD D5)
(To what extent do you agree or disagree with the following statements?)
Strong trade unions are bad for America's economy.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6328/vshow}{GSS Data Explorer page for \code{unionsbd}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 63 \tab 204 \tab 532 \tab 472 \tab 158 \tab 39 \tab 9 \tab 2867 \cr
   Total \tab 1390 \tab 63 \tab 204 \tab 532 \tab 472 \tab 158 \tab 39 \tab 9 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
