% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{united}
\alias{united}
\title{Americans united or divided on values}
\source{
General Social Survey https://gss.norc.org
}
\description{
united
}
\details{
Question A. There is a lot of discussion today about whether Americans are divided or united. Some say that Americans are united and in agreement about the most important values. Others think that Americans are greatly divided when it comes to the most important values. What is your view about this?
}
\section{Values}{

\itemize{
\item \code{1} american are united
\item \code{2} american are divided
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1380/vshow}{GSS Data Explorer page for \code{united}}.
}

\seealso{
Other Culture: 
\code{\link{bigband}},
\code{\link{blugrass}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{obtohelp}},
\code{\link{opoutcme}}
}
\concept{Culture}
\concept{Multiculturalism}
\concept{Values}
\keyword{variable}
