% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{unjobexp}
\alias{unjobexp}
\title{Rank of how to spend time when unemployed}
\source{
General Social Survey https://gss.norc.org
}
\description{
unjobexp
}
\details{
Question Suppose you were unemployed and couldn't find a job. Which of the following problems would be the worst?  Please write '1' in the box beside the worst thing, write '2' in the box beside the next worst thing, and '3' beside the third worst thing. Leave the other boxes blank.
F. Loss of job experience
}
\section{Values}{

\itemize{
\item \code{1} worst
\item \code{2} 2nd worst
\item \code{3} 3rd worst
\item \code{4} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4521/vshow}{GSS Data Explorer page for \code{unjobexp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab worst \tab 2nd worst \tab 3rd worst \tab not chosen \tab don't know \tab no answer \tab Total \cr
   1989 \tab 84 \tab 24 \tab 93 \tab 124 \tab 864 \tab 56 \tab 292 \tab 1537 \cr
   Total \tab 84 \tab 24 \tab 93 \tab 124 \tab 864 \tab 56 \tab 292 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{econsup1}},
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{unfamten}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{worecsup}}
}
\concept{Work Orientation}
\keyword{variable}
