% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{unmanrel}
\alias{unmanrel}
\title{Relations between management and workers}
\source{
General Social Survey https://gss.norc.org
}
\description{
unmanrel
}
\details{
Question In general how would you describe relations in your work place between management and employees?  Would you say they are very good, quite good, neither good nor bad, quite bad, or very bad?
}
\section{Values}{

\itemize{
\item \verb{[1]} very
\item \verb{[2]} quite good
\item \verb{[3]} neither good nor bad
\item \verb{[4]} quite bad
\item \verb{[5]} very bad
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1366/vshow}{GSS Data Explorer page for \code{unmanrel}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very \tab quite good \tab neither good nor bad \tab quite bad \tab very bad \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 289 \tab 314 \tab 208 \tab 40 \tab 24 \tab 8 \tab 29 \tab 1517 \cr
   Total \tab 605 \tab 289 \tab 314 \tab 208 \tab 40 \tab 24 \tab 8 \tab 29 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{jobvshme}},
\code{\link{schoolng}},
\code{\link{unvote}}
}
\concept{Work Organizations}
\keyword{variable}
