% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{unpeople}
\alias{unpeople}
\title{Rank of no people contact when unemployed}
\source{
General Social Survey https://gss.norc.org
}
\description{
unpeople
}
\details{
Question Suppose you were unemployed and couldn't find a job. Which of the following problems would be the worst?  Please write '1' in the box beside the worst thing, write '2' in the box beside the next worst thing, and '3' beside the third worst thing. Leave the other boxes blank.
A. Lack of contact with people at work
}
\section{Values}{

\itemize{
\item \verb{[1]} worst
\item \verb{[2]} 2nd worst
\item \verb{[3]} 3rd worst
\item \verb{[4]} not chosen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4516/vshow}{GSS Data Explorer page for \code{unpeople}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab worst \tab 2nd worst \tab 3rd worst \tab not chosen \tab don't know \tab no answer \tab Total \cr
   1989 \tab 84 \tab 72 \tab 91 \tab 135 \tab 815 \tab 56 \tab 284 \tab 1537 \cr
   Total \tab 84 \tab 72 \tab 91 \tab 135 \tab 815 \tab 56 \tab 284 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{econsup1}},
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{worecsup}}
}
\concept{Work Orientation}
\keyword{variable}
