% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{unplgrsn}
\alias{unplgrsn}
\title{Reason r reduced time spent on internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
unplgrsn
}
\details{
Question What is the main reason why you deliberately reduced the time you spent on the Internet?
}
\section{Values}{

\itemize{
\item \verb{[1]} i had concerns about data security or privacy issues
\item \verb{[2]} i felt it negatively impacted my physical or mental wellbeing
\item \verb{[3]} i felt overwhelmed by the amount of information
\item \verb{[4]} i was asked to spend more time in person with family or friends
\item \verb{[5]} i wanted more time for offline activities
\item \verb{[6]} i couldn't afford to use the internet
\item \verb{[7]} other reasons
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8179/vshow}{GSS Data Explorer page for \code{unplgrsn}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab i had concerns about data security or privacy issues \tab i felt it negatively impacted my physical or mental wellbeing \tab i felt overwhelmed by the amount of information \tab i was asked to spend more time in person with family or friends \tab i wanted more time for offline activities \tab i couldn't afford to use the internet \tab other reasons \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 80 \tab 273 \tab 65 \tab 91 \tab 327 \tab 12 \tab 70 \tab 41 \tab 2333 \tab 2 \tab 15 \tab 3309 \cr
   Total \tab 0 \tab 80 \tab 273 \tab 65 \tab 91 \tab 327 \tab 12 \tab 70 \tab 41 \tab 2333 \tab 2 \tab 15 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{inthome}},
\code{\link{intliteracy}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
