% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{unplug}
\alias{unplug}
\title{Amount r reduced time spent on internet in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
unplug
}
\details{
Question In the last 12 months, how often, if at all, did you deliberately reduce the time you spent on the internet?
}
\section{Values}{

\itemize{
\item \verb{[1]} every day or almost every day
\item \verb{[2]} several times a week
\item \verb{[3]} several times a month
\item \verb{[4]} several times a year
\item \verb{[5]} less often
\item \verb{[6]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8178/vshow}{GSS Data Explorer page for \code{unplug}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab every day or almost every day \tab several times a week \tab several times a month \tab several times a year \tab less often \tab never \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 110 \tab 289 \tab 218 \tab 138 \tab 208 \tab 489 \tab 108 \tab 1731 \tab 5 \tab 13 \tab 3309 \cr
   Total \tab 0 \tab 110 \tab 289 \tab 218 \tab 138 \tab 208 \tab 489 \tab 108 \tab 1731 \tab 5 \tab 13 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{inthome}},
\code{\link{intliteracy}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplgrsn}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
