% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{unpower}
\alias{unpower}
\title{Unions in this country have too little power}
\source{
General Social Survey https://gss.norc.org
}
\description{
unpower
}
\details{
Question To what extent do you agree or disagree with the following statements?
Please tell me whether you strongly agree, agree, disagree, or strongly disagree:
C. Unions in this country have too little power…
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1265/vshow}{GSS Data Explorer page for \code{unpower}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab - \tab 87 \tab 336 \tab 721 \tab 203 \tab 165 \tab 5 \tab - \tab 1517 \cr
   Total \tab 0 \tab 87 \tab 336 \tab 721 \tab 203 \tab 165 \tab 5 \tab 0 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{afailure}},
\code{\link{automatn}},
\code{\link{baseofop}},
\code{\link{chngeorg}},
\code{\link{cowrkrel}},
\code{\link{drugtest}},
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{futpromo}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{helporg}},
\code{\link{idecide}},
\code{\link{imatter}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobcntrl}},
\code{\link{jobjudge}},
\code{\link{jobvshme}},
\code{\link{jobyears}},
\code{\link{learning}},
\code{\link{levels}()},
\code{\link{lottosay}},
\code{\link{ltrsref}},
\code{\link{merged}},
\code{\link{monitred}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{offsup}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{orgfin}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{othpay}},
\code{\link{outofbiz}},
\code{\link{pessimst}},
\code{\link{physical}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{proudorg}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{racpromo}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{reorg}},
\code{\link{samejob}},
\code{\link{samevals}},
\code{\link{satself}},
\code{\link{schoolng}},
\code{\link{sexpromo}},
\code{\link{skiltest}},
\code{\link{slfrspct}},
\code{\link{splocnum}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{totunemp}},
\code{\link{training}},
\code{\link{unmanrel}},
\code{\link{unprog}},
\code{\link{unvote}},
\code{\link{wherewrk}},
\code{\link{wrkindep}},
\code{\link{wrkmuch}},
\code{\link{wrkwell}},
\code{\link{wrkyears}},
\code{\link{youmoney}},
\code{\link{yousup}}

Other Labor Unions: 
\code{\link{memunion}},
\code{\link{sallabor}},
\code{\link{strngun}},
\code{\link{trdunion}},
\code{\link{union}()},
\code{\link{unionsok}},
\code{\link{unprog}},
\code{\link{unvote}}
}
\concept{Labor Unions}
\concept{Work Organizations}
\keyword{variable}
