% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_15.R
\name{unrelat}
\alias{unrelat}
\title{Number in household not related}
\source{
General Social Survey https://gss.norc.org
}
\description{
unrelat
}
\details{
Question Is everyone in the household related to you in some way?  A. IF NO:  How many persons in the household are not related to you in any way?
As of 2024, UNRELAT has been recoded to only account for 0 to 2+ people. This has been retroactively applied to all previous years of UNRELAT. The original UNRELAT containing data from 1972 to 2022 has been renamed UNRELAT_7222.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0
\item \verb{[1]} 1
\item \verb{[2]} 2 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/102/vshow}{GSS Data Explorer page for \code{unrelat}}.

Counts by year:\tabular{lllllllll}{
   year \tab 0 \tab 1 \tab 2 or more \tab iap \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1388 \tab 44 \tab 25 \tab 153 \tab 3 \tab - \tab - \tab 1613 \cr
   1973 \tab 1273 \tab 50 \tab 27 \tab 154 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1244 \tab 52 \tab 19 \tab 168 \tab 1 \tab - \tab - \tab 1484 \cr
   1975 \tab 1223 \tab 49 \tab 18 \tab 200 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1196 \tab 50 \tab 17 \tab 235 \tab 1 \tab - \tab - \tab 1499 \cr
   1977 \tab 1208 \tab 45 \tab 16 \tab 261 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1164 \tab 53 \tab 10 \tab 296 \tab 9 \tab - \tab - \tab 1532 \cr
   1980 \tab 1094 \tab 58 \tab 22 \tab 288 \tab 6 \tab - \tab - \tab 1468 \cr
   1982 \tab 1345 \tab 78 \tab 18 \tab 413 \tab 6 \tab - \tab - \tab 1860 \cr
   1983 \tab 1199 \tab 63 \tab 15 \tab 309 \tab 13 \tab - \tab - \tab 1599 \cr
   1984 \tab 1048 \tab 67 \tab 27 \tab 330 \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 1046 \tab 80 \tab 26 \tab 343 \tab 39 \tab - \tab - \tab 1534 \cr
   1986 \tab 1048 \tab 67 \tab 27 \tab 312 \tab 16 \tab - \tab - \tab 1470 \cr
   1987 \tab 1232 \tab 90 \tab 53 \tab 420 \tab 24 \tab - \tab - \tab 1819 \cr
   1988 \tab 1034 \tab 94 \tab 24 \tab 328 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 1073 \tab 101 \tab 23 \tab 327 \tab 13 \tab - \tab - \tab 1537 \cr
   1990 \tab 935 \tab 76 \tab 14 \tab 330 \tab 17 \tab - \tab - \tab 1372 \cr
   1991 \tab 1009 \tab 76 \tab 32 \tab 378 \tab 22 \tab - \tab - \tab 1517 \cr
   1993 \tab 1040 \tab 90 \tab 34 \tab 436 \tab 6 \tab - \tab - \tab 1606 \cr
   1994 \tab 1846 \tab 163 \tab 46 \tab 903 \tab 34 \tab - \tab - \tab 2992 \cr
   1996 \tab 1611 \tab 201 \tab 73 \tab 949 \tab 70 \tab - \tab - \tab 2904 \cr
   1998 \tab 1609 \tab 166 \tab 73 \tab 944 \tab 40 \tab - \tab - \tab 2832 \cr
   2000 \tab 1640 \tab 194 \tab 64 \tab 917 \tab 2 \tab - \tab - \tab 2817 \cr
   2002 \tab 1647 \tab 168 \tab 68 \tab 881 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 1744 \tab 166 \tab 41 \tab 861 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2692 \tab 231 \tab 98 \tab 1483 \tab 6 \tab - \tab - \tab 4510 \cr
   2008 \tab 1245 \tab 104 \tab 36 \tab 637 \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 1135 \tab 137 \tab 48 \tab 724 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1156 \tab 133 \tab 53 \tab 632 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1456 \tab 167 \tab 70 \tab 844 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 1587 \tab 175 \tab 75 \tab 1025 \tab 5 \tab - \tab - \tab 2867 \cr
   2018 \tab 1315 \tab 157 \tab 50 \tab 826 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 2285 \tab 332 \tab 82 \tab 1318 \tab - \tab 1 \tab 14 \tab 4032 \cr
   2022 \tab 1827 \tab 283 \tab 90 \tab 1339 \tab - \tab - \tab 5 \tab 3544 \cr
   2024 \tab 1670 \tab 249 \tab 71 \tab 1312 \tab - \tab - \tab 7 \tab 3309 \cr
   Total \tab 48264 \tab 4309 \tab 1485 \tab 21276 \tab 338 \tab 1 \tab 26 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{granborn}},
\code{\link{hompop}},
\code{\link{income}},
\code{\link{preteen}},
\code{\link{rincome}},
\code{\link{teens}}

Other Household: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{hompop}},
\code{\link{preteen}},
\code{\link{teens}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
