% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{unrghts}
\alias{unrghts}
\title{Un should intervene if human right is seriously viola}
\source{
General Social Survey https://gss.norc.org
}
\description{
unrghts
}
\details{
Question Which of these two statements comes closer to your view?
}
\section{Values}{

\itemize{
\item \verb{[1]} un should intervene
\item \verb{[2]} un should not intervene
\item \verb{[3]} dont know what the un is
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4951/vshow}{GSS Data Explorer page for \code{unrghts}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab un should intervene \tab un should not intervene \tab dont know what the un is \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 1067 \tab 271 \tab 61 \tab 51 \tab 22 \tab 2812 \cr
   Total \tab 1340 \tab 1067 \tab 271 \tab 61 \tab 51 \tab 22 \tab 2812 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{servepeo}}
}
\concept{ISSP Citizenship}
\keyword{variable}
