% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{upbreath}
\alias{upbreath}
\title{Why short breath at the top of a mountain}
\source{
General Social Survey https://gss.norc.org
}
\description{
upbreath
}
\details{
Question For which reason may people experience shortness of breath more quickly at the top of a mountain than along a seashore?
}
\section{Values}{

\itemize{
\item \verb{[1]} a slower pulse rate
\item \verb{[2]} a greater gravitational force on the body
\item \verb{[3]} a lower percent of oxygen in the blood
\item \verb{[4]} a faster heartbeat
\item \verb{[5]} a slower circulation of blood
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3481/vshow}{GSS Data Explorer page for \code{upbreath}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a slower pulse rate \tab a greater gravitational force on the body \tab a lower percent of oxygen in the blood \tab a faster heartbeat \tab a slower circulation of blood \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 24 \tab 181 \tab 1016 \tab 67 \tab 58 \tab 137 \tab 22 \tab 2023 \cr
   Total \tab 518 \tab 24 \tab 181 \tab 1016 \tab 67 \tab 58 \tab 137 \tab 22 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
