% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{upgrade}
\alias{upgrade}
\title{Refuse vols with no basic skills?}
\source{
General Social Survey https://gss.norc.org
}
\description{
upgrade
}
\details{
Question Many people who want to volunteer for service in the armed forces Do not have the necessary basic skills like reading, writing and arithmetic. Do you think the armed forces should refuse to accept such volunteers, or should they accept them and give them the necessary education?
}
\section{Values}{

\itemize{
\item \code{1} refuse to accept
\item \code{2} accept & educate
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/820/vshow}{GSS Data Explorer page for \code{upgrade}}.
}

\seealso{
Other Core: 
\code{\link{defwrkev}},
\code{\link{guerilla}},
\code{\link{jobtrain}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{vetaid}}

Other Military: 
\code{\link{defwrkev}},
\code{\link{jobtrain}}
}
\concept{Core}
\concept{Military}
\keyword{variable}
