% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{upload}
\alias{upload}
\title{Does r know how to upload files to www}
\source{
General Social Survey https://gss.norc.org
}
\description{
upload
}
\details{
Question Now we'd like to ask you a few questions about terms that people use to talk about the Internet and World Wide Web.
D. Do you know how to send a file that is on your computer's hard drive to someone using another computer?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2514/vshow}{GSS Data Explorer page for \code{upload}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 457 \tab 205 \tab 80 \tab - \tab 2817 \cr
   2004 \tab 992 \tab 1360 \tab 437 \tab 20 \tab 3 \tab 2812 \cr
   Total \tab 3067 \tab 1817 \tab 642 \tab 100 \tab 3 \tab 5629 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{srcheng2}}
}
\concept{Information Society}
\keyword{variable}
