% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{upset}
\alias{upset}
\title{How often r's friends or family were angry with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
upset
}
\details{
Question Thinking about the important people in your life, such as your spouse or partner, your family members, or close friends, how often in the past 4 weeks did any of these people act angry or upset with you?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} rarely
\item \code{3} sometimes
\item \code{4} often
\item \code{5} very often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7260/vshow}{GSS Data Explorer page for \code{upset}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab rarely \tab sometimes \tab often \tab very often \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 454 \tab 444 \tab 211 \tab 42 \tab 19 \tab 1 \tab 2 \tab 2348 \cr
   Total \tab 1175 \tab 454 \tab 444 \tab 211 \tab 42 \tab 19 \tab 1 \tab 2 \tab 2348 \cr
}
}

\keyword{variable}
