% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_455.R
\name{upset1}
\alias{upset1}
\title{Help for marital problem #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
upset1
}
\details{
Question Suppose you were very upset about a problem with your husband, wife or partner and hadn't been able to work it out with them. Who would you turn to for help? Even if you are not married or have no partner, what would you do?
A. Who would you turn to first for help?
B. Who would you turn to second?
PLEASE CIRCLE ONLY ONE FIRST CHOICE NUMBER AND ONE SECOND CHOICE NUMBER.
A. First Choice
}
\section{Values}{

\itemize{
\item \verb{[1]} spouse
\item \verb{[2]} mother
\item \verb{[3]} father
\item \verb{[4]} daughter
\item \verb{[5]} son
\item \verb{[6]} sister
\item \verb{[7]} brother
\item \verb{[8]} other relative
\item \verb{[9]} closest friend
\item \verb{[10]} other friend
\item \verb{[11]} neighbor
\item \verb{[12]} coworker
\item \verb{[13]} clergy
\item \verb{[14]} family doctor
\item \verb{[15]} prof counselor
\item \verb{[16]} other
\item \verb{[17]} no one
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4175/vshow}{GSS Data Explorer page for \code{upset1}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab neighbor \tab coworker \tab clergy \tab family doctor \tab prof counselor \tab other \tab no one \tab no answer \tab Total \cr
   1986 \tab 42 \tab 101 \tab 167 \tab 35 \tab 90 \tab 52 \tab 114 \tab 45 \tab 52 \tab 381 \tab 18 \tab 7 \tab 8 \tab 187 \tab 15 \tab 55 \tab 3 \tab 67 \tab 31 \tab 1470 \cr
   Total \tab 42 \tab 101 \tab 167 \tab 35 \tab 90 \tab 52 \tab 114 \tab 45 \tab 52 \tab 381 \tab 18 \tab 7 \tab 8 \tab 187 \tab 15 \tab 55 \tab 3 \tab 67 \tab 31 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1}},
\code{\link{borrow2}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{down1}},
\code{\link{down2}},
\code{\link{sick1}},
\code{\link{sick2}},
\code{\link{upset2}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
