% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{urbantech}
\alias{urbantech}
\title{Are cities or countryside benefiting more from digital tech}
\source{
General Social Survey https://gss.norc.org
}
\description{
urbantech
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} people who live in the countryside are benefitting more than people who live in cities
\item \code{2} people who live in the countryside and people who live in cities are benefitting equally
\item \code{3} people who live in cities are benefitting more than people who live in the countryside
\item \code{4} neither people from the countryside nor from cities are benefitting
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
