% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{usangry}
\alias{usangry}
\title{Thinking us how often feel angry}
\source{
General Social Survey https://gss.norc.org
}
\description{
usangry
}
\details{
Question Now I want to ask some questions about how you feel when you think about this country, the United States. When you think about the United States:
E. When you think about the United States, how often do you feel ANGRY? Do you feel angry:
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} most of the time
\item \verb{[3]} some of the time
\item \verb{[4]} rarely
\item \verb{[5]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1793/vshow}{GSS Data Explorer page for \code{usangry}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab most of the time \tab some of the time \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2199 \tab 10 \tab 41 \tab 299 \tab 246 \tab 99 \tab 8 \tab 2 \tab 2904 \cr
   Total \tab 2199 \tab 10 \tab 41 \tab 299 \tab 246 \tab 99 \tab 8 \tab 2 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{usenthus}},
\code{\link{ushopefl}},
\code{\link{usupset}}
}
\concept{Emotions}
\keyword{variable}
