% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_256.R
\name{usasn}
\alias{usasn}
\title{What percentage of population is asian}
\source{
General Social Survey https://gss.norc.org
}
\description{
usasn
}
\details{
Question Just your best guess–what percentage of the United States population is each group? (INSTRUCTION: PERCENTAGES DO NOT HAVE TO ADD TO 100pct AND THE LISTED GROUPS MAY OVERLAP. ENCOURAGE ESTIMATES FOR ALL GROUPS, BUT DISCOURAGE RESPONDENTS REVISING THEIR ESTIMATES.)
E. Asian Americans?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2314/vshow}{GSS Data Explorer page for \code{usasn}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 18 \tab 19 \tab 2 \tab 20 \tab 22 \tab 25 \tab 3 \tab 30 \tab 33 \tab 35 \tab 37 \tab 38 \tab 4 \tab 40 \tab 45 \tab 5 \tab 50 \tab 55 \tab 6 \tab 60 \tab 63 \tab 65 \tab 7 \tab 70 \tab 75 \tab 8 \tab 80 \tab 85 \tab 9 \tab 90 \tab 95 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 5 \tab 15 \tab 239 \tab 2 \tab 1 \tab 4 \tab 3 \tab 2 \tab 116 \tab 1 \tab 4 \tab 1 \tab 44 \tab 162 \tab 1 \tab 58 \tab 39 \tab 103 \tab 1 \tab 10 \tab 1 \tab 1 \tab 18 \tab 56 \tab 6 \tab 213 \tab 45 \tab 4 \tab 11 \tab 15 \tab 1 \tab 4 \tab 20 \tab 7 \tab 3 \tab 30 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 111 \tab 31 \tab 2817 \cr
   Total \tab 1419 \tab 5 \tab 15 \tab 239 \tab 2 \tab 1 \tab 4 \tab 3 \tab 2 \tab 116 \tab 1 \tab 4 \tab 1 \tab 44 \tab 162 \tab 1 \tab 58 \tab 39 \tab 103 \tab 1 \tab 10 \tab 1 \tab 1 \tab 18 \tab 56 \tab 6 \tab 213 \tab 45 \tab 4 \tab 11 \tab 15 \tab 1 \tab 4 \tab 20 \tab 7 \tab 3 \tab 30 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 111 \tab 31 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkchng}},
\code{\link{hispchng}},
\code{\link{jewschng}},
\code{\link{usamind}},
\code{\link{usblk}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{whtchng}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
