% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{uscitzn}
\alias{uscitzn}
\title{Is r us citizen}
\source{
General Social Survey https://gss.norc.org
}
\description{
uscitzn
}
\details{
Question If not born in US, is R a US citizen
As of 2024, USCITZN has been recoded to simplify responses. It is now a synthetic variable which represents the original USCITZN, as well as the experimental USCITZNV and USCITZNNV. This recode is retroactive, covering all previous years of USCITZN. The original data have been preserved in USCITZN_0822. USCITZNV and USCITZNNV will not be made available in 2024.
}
\section{Values}{

\itemize{
\item \verb{[1]} a u.s. citizen
\item \verb{[2]} not a u.s. citizen
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5283/vshow}{GSS Data Explorer page for \code{uscitzn}}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab a u.s. citizen \tab not a u.s. citizen \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab - \tab 126 \tab 133 \tab 1 \tab 1758 \tab 5 \tab - \tab 2023 \cr
   2010 \tab - \tab 122 \tab 122 \tab 3 \tab 1795 \tab 2 \tab - \tab 2044 \cr
   2012 \tab - \tab 144 \tab 123 \tab 2 \tab 1700 \tab 5 \tab - \tab 1974 \cr
   2014 \tab - \tab 194 \tab 157 \tab - \tab 2184 \tab 3 \tab - \tab 2538 \cr
   2016 \tab - \tab 197 \tab 148 \tab 5 \tab 2507 \tab 10 \tab - \tab 2867 \cr
   2018 \tab - \tab 167 \tab 131 \tab 1 \tab 2045 \tab 4 \tab - \tab 2348 \cr
   2021 \tab - \tab 305 \tab 156 \tab - \tab 3562 \tab 1 \tab 8 \tab 4032 \cr
   2022 \tab - \tab 284 \tab 161 \tab 4 \tab 3082 \tab 10 \tab 3 \tab 3544 \cr
   2024 \tab - \tab 263 \tab 172 \tab 3 \tab 2859 \tab 2 \tab 10 \tab 3309 \cr
   Total \tab 0 \tab 1802 \tab 1303 \tab 19 \tab 21492 \tab 42 \tab 21 \tab 24679 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{fucitzn}}

Other Citizenship: 
\code{\link{fucitzn}}
}
\concept{Citizenship}
\concept{Ethnicity}
\keyword{variable}
