% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_691.R
\name{uscitzn_0822}
\alias{uscitzn_0822}
\title{If not born in us, is r a us citizen}
\source{
General Social Survey https://gss.norc.org
}
\description{
uscitzn_0822
}
\details{
Question If not born in US, is R a US citizen
As of 2024, USCITZN has been recoded to simplify responses. It is now a synthetic variable which represents the original USCITZN, as well as the experimental USCITZNV and USCITZNNV. This recode is retroactive, covering all previous years of USCITZN. The original data have been preserved in USCITZN_0822. USCITZNV and USCITZNNV will not be made available in 2024.
}
\section{Values}{

\itemize{
\item \verb{[1]} a u.s. citizen
\item \verb{[2]} not a u.s. citizen
\item \verb{[3]} a u.s. citizen born in puerto rico, the u.s. virgin islands, or the northern marianas islands (if volunteered)
\item \verb{[4]} born outside of the u.s. to parents who were u.s. citizens at that time (if volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2010, 2012, 2014, 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8181/vshow}{GSS Data Explorer page for \code{uscitzn_0822}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab a u.s. citizen \tab not a u.s. citizen \tab a u.s. citizen born in puerto rico, the u.s. virgin islands, or the northern marianas islands (if volunteered) \tab born outside of the u.s. to parents who were u.s. citizens at that time (if volunteered) \tab don't know \tab iap \tab no answer \tab Total \cr
   2008 \tab - \tab 118 \tab 133 \tab 2 \tab 6 \tab 1 \tab 1758 \tab 5 \tab 2023 \cr
   2010 \tab - \tab 116 \tab 122 \tab 3 \tab 3 \tab 3 \tab 1795 \tab 2 \tab 2044 \cr
   2012 \tab - \tab 141 \tab 123 \tab 1 \tab 2 \tab 2 \tab 1700 \tab 5 \tab 1974 \cr
   2014 \tab - \tab 187 \tab 157 \tab 3 \tab 4 \tab - \tab 2184 \tab 3 \tab 2538 \cr
   2016 \tab - \tab 180 \tab 148 \tab 8 \tab 9 \tab 5 \tab 2507 \tab 10 \tab 2867 \cr
   2018 \tab - \tab 163 \tab 131 \tab 2 \tab 2 \tab 1 \tab 2045 \tab 4 \tab 2348 \cr
   2022 \tab - \tab 151 \tab 92 \tab 4 \tab 6 \tab 4 \tab 3277 \tab 10 \tab 3544 \cr
   Total \tab 0 \tab 1056 \tab 906 \tab 23 \tab 32 \tab 16 \tab 15266 \tab 39 \tab 17338 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{hispanic_0022}},
\code{\link{racecen1_0022}},
\code{\link{racecen2_0022}},
\code{\link{racecen3_0022}}
}
\concept{Citizenship}
\concept{Ethnicity}
\keyword{variable}
