% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_687.R
\name{uscitznv_2122}
\alias{uscitznv_2122}
\title{Uscitznv_2122}
\source{
General Social Survey https://gss.norc.org
}
\description{
uscitznv_2122
}
\details{
Question uscitznv_2122
}
\section{Values}{

\itemize{
\item \code{1} a u.s. citizen
\item \code{2} not a u.s. citizen
\item \code{3} a u.s. citizen born in puerto rico, the u.s. virgin islands, or the northern marianas islands (phone mode only: v
\item \code{4} born outside of the u.s. to parents who were u.s. citizens at that time (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org}{GSS Data Explorer page for \code{uscitznv_2122}}.
}

\keyword{variable}
