% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_29.R
\name{usclass1}
\alias{usclass1}
\title{Traditional class divisions still remain}
\source{
General Social Survey https://gss.norc.org
}
\description{
usclass1
}
\details{
Question Here are different opinions about social differences in this country. Please tell me for each one whether you strongly agree, somewhat, agree, somewhat disagree, or strongly disagree. READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.
A. In the United States traditional divisions between owners and workers still remain. A person's social standing depends upon whether he/she belongs to the upper or lower class.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} somewhat agree
\item \verb{[3]} somewhat disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/233/vshow}{GSS Data Explorer page for \code{usclass1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 230 \tab 750 \tab 323 \tab 92 \tab 51 \tab 27 \tab 1473 \cr
   Total \tab 0 \tab 230 \tab 750 \tab 323 \tab 92 \tab 51 \tab 27 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}
}
\concept{Core}
\keyword{variable}
