% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_29.R
\name{usclass5}
\alias{usclass5}
\title{Personal income not determined by work}
\source{
General Social Survey https://gss.norc.org
}
\description{
usclass5
}
\details{
Question Here are different opinions about social differences in this country. Please tell me for each one whether you strongly agree, somewhat, agree, somewhat disagree, or strongly disagree. READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.
E. Personal income should not be determined solely by one's work. Rather, everybody should get what he/she needs to provide a decent life for his/her family.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} somewhat agree
\item \verb{[3]} somewhat disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/237/vshow}{GSS Data Explorer page for \code{usclass5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 163 \tab 316 \tab 492 \tab 444 \tab 29 \tab 29 \tab 1473 \cr
   Total \tab 0 \tab 163 \tab 316 \tab 492 \tab 444 \tab 29 \tab 29 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}
}
\concept{Core}
\keyword{variable}
