% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_29.R
\name{usclass8}
\alias{usclass8}
\title{Social differences are acceptable}
\source{
General Social Survey https://gss.norc.org
}
\description{
usclass8
}
\details{
Question Here are different opinions about social differences in this country. Please tell me for each one whether you strongly agree, somewhat, agree, somewhat disagree, or strongly disagree. READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.
H. All in all, I think social differences in this country are justified.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/240/vshow}{GSS Data Explorer page for \code{usclass8}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 120 \tab 631 \tab 474 \tab 164 \tab 58 \tab 26 \tab 1473 \cr
   Total \tab 0 \tab 120 \tab 631 \tab 474 \tab 164 \tab 58 \tab 26 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}}
}
\concept{Core}
\keyword{variable}
