% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{usdoenuf}
\alias{usdoenuf}
\title{U.s. protect environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
usdoenuf
}
\details{
Question C. Some countries are doing more to protect the world environment than other countries are. in general, do you think that America is doing...
}
\section{Values}{

\itemize{
\item \verb{[1]} more than enough
\item \verb{[2]} about the right amount, or
\item \verb{[3]} too little?
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4768/vshow}{GSS Data Explorer page for \code{usdoenuf}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab more than enough \tab about the right amount, or \tab too little? \tab don't know \tab no answer \tab Total \cr
   2010 \tab 614 \tab 129 \tab 473 \tab 693 \tab 130 \tab 5 \tab 2044 \cr
   Total \tab 614 \tab 129 \tab 473 \tab 693 \tab 130 \tab 5 \tab 2044 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}}
}
\concept{ISSP Environment}
\keyword{variable}
