% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{usedproz}
\alias{usedproz}
\title{Have you ever taken prozac}
\source{
General Social Survey https://gss.norc.org
}
\description{
usedproz
}
\details{
Question Next, I''d like to ask you some questions about the medication called Prozac.
B. Have you, yourself, ever taken Prozac?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2221/vshow}{GSS Data Explorer page for \code{usedproz}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1635 \tab 97 \tab 1078 \tab 7 \tab 15 \tab 2832 \cr
   2006 \tab 3238 \tab 132 \tab 1123 \tab 2 \tab 15 \tab 4510 \cr
   Total \tab 4873 \tab 229 \tab 2201 \tab 9 \tab 30 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}}
}
\concept{Mental Health}
\keyword{variable}
