% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{usedup}
\alias{usedup}
\title{How often during past month r felt used up}
\source{
General Social Survey https://gss.norc.org
}
\description{
usedup
}
\details{
Question How often during the past month have you felt used up at the end of the day?
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} often
\item \code{3} sometimes
\item \code{4} rarely
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2830/vshow}{GSS Data Explorer page for \code{usedup}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very often \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 333 \tab 419 \tab 597 \tab 301 \tab 123 \tab 7 \tab 16 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 330 \tab 401 \tab 587 \tab 280 \tab 123 \tab 3 \tab 10 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 200 \tab 279 \tab 382 \tab 210 \tab 89 \tab 3 \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 187 \tab 298 \tab 423 \tab 245 \tab 88 \tab 4 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 207 \tab 312 \tab 503 \tab 279 \tab 113 \tab 1 \tab 33 \tab - \tab 2348 \cr
   2022 \tab 1557 \tab 333 \tab 422 \tab 687 \tab 349 \tab 161 \tab 5 \tab 13 \tab 17 \tab 3544 \cr
   Total \tab 8347 \tab 1590 \tab 2131 \tab 3179 \tab 1664 \tab 697 \tab 23 \tab 101 \tab 17 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{backpain}},
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{strredpg}}
}
\concept{Quality of Working Life}
\keyword{variable}
