% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_539.R
\name{usemedia}
\alias{usemedia}
\title{Contacted in the media to express view}
\source{
General Social Survey https://gss.norc.org
}
\description{
usemedia
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the past year, whether you have done it in the more distant past, whether you have not done it but might do it, or have not done it and would never, under any circumstances, do it.
G. Contacted or appeared in the media to express your views
}
\section{Values}{

\itemize{
\item \code{1} have done it in the past yr
\item \code{2} have done it in the more distant past
\item \code{3} have not done it but might do it
\item \code{4} have not done it and would never do it
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4926/vshow}{GSS Data Explorer page for \code{usemedia}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab have done it in the past yr \tab have done it in the more distant past \tab have not done it but might do it \tab have not done it and would never do it \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 75 \tab 137 \tab 671 \tab 583 \tab 2 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 47 \tab 93 \tab 482 \tab 618 \tab 22 \tab 2 \tab 2538 \cr
   Total \tab 2614 \tab 122 \tab 230 \tab 1153 \tab 1201 \tab 24 \tab 6 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{cntctgov}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{polfunds}},
\code{\link{racmeet}},
\code{\link{revmeet}},
\code{\link{signdpet}}
}
\concept{ISSP Citizenship}
\keyword{variable}
