% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_242.R
\name{usepsyc1}
\alias{usepsyc1}
\title{Will take medicine because personal trouble}
\source{
General Social Survey https://gss.norc.org
}
\description{
usepsyc1
}
\details{
Question How likely would you be to take doctor-prescribed psychiatric medication in the following situations?
A. ...because you were having trouble in your personal life.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2175/vshow}{GSS Data Explorer page for \code{usepsyc1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 116 \tab 205 \tab 155 \tab 252 \tab 610 \tab 46 \tab 3 \tab 2832 \cr
   2006 \tab 3073 \tab 158 \tab 260 \tab 241 \tab 258 \tab 474 \tab 32 \tab 14 \tab 4510 \cr
   Total \tab 4518 \tab 274 \tab 465 \tab 396 \tab 510 \tab 1084 \tab 78 \tab 17 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}}
}
\concept{Mental Health}
\keyword{variable}
