% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_242.R
\name{usepsyc2}
\alias{usepsyc2}
\title{Will take medicine because cant cope w stress}
\source{
General Social Survey https://gss.norc.org
}
\description{
usepsyc2
}
\details{
Question How likely would you be to take doctor-prescribed psychiatric medication in the following situations?
B. ...because you didn''t know how to cope anymore with the stresses of life.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} mixed
\item \code{4} somewht unlikely
\item \code{5} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2176/vshow}{GSS Data Explorer page for \code{usepsyc2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 168 \tab 322 \tab 180 \tab 241 \tab 427 \tab 45 \tab 4 \tab 2832 \cr
   2006 \tab 3073 \tab 243 \tab 415 \tab 229 \tab 185 \tab 310 \tab 42 \tab 13 \tab 4510 \cr
   Total \tab 4518 \tab 411 \tab 737 \tab 409 \tab 426 \tab 737 \tab 87 \tab 17 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{usepsyc1}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}}
}
\concept{Mental Health}
\keyword{variable}
