% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_242.R
\name{usepsyc4}
\alias{usepsyc4}
\title{Will take medicine because intense fear}
\source{
General Social Survey https://gss.norc.org
}
\description{
usepsyc4
}
\details{
Question How likely would you be to take doctor-prescribed psychiatric medication in the following situations?
D. ...for no apparent reason, you were having periods of intense fear in which you were trembling, sweating, feeling dizzy, and feared losing control or going crazy.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} mixed
\item \verb{[4]} somewht unlikely
\item \verb{[5]} very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2178/vshow}{GSS Data Explorer page for \code{usepsyc4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab mixed \tab somewht unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 351 \tab 391 \tab 182 \tab 135 \tab 267 \tab 58 \tab 3 \tab 2832 \cr
   2006 \tab 3073 \tab 505 \tab 384 \tab 178 \tab 116 \tab 205 \tab 35 \tab 14 \tab 4510 \cr
   Total \tab 4518 \tab 856 \tab 775 \tab 360 \tab 251 \tab 472 \tab 93 \tab 17 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}}
}
\concept{Mental Health}
\keyword{variable}
