% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{usewww}
\alias{usewww}
\title{R use www other than email}
\source{
General Social Survey https://gss.norc.org
}
\description{
usewww
}
\details{
Question Other than for e-mail, do you ever use the Internet or World Wide Web? (INTERVIEWER: TREAT THESE AS THE SAME)
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2383/vshow}{GSS Data Explorer page for \code{usewww}}.
}

\seealso{
Other Information Society: 
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{drctlink}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{srcheng}},
\code{\link{typeurl}},
\code{\link{wwwhr}},
\code{\link{wwwmin}}

Other Computers And Internet: 
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{drctlink}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{srcheng}},
\code{\link{typeurl}},
\code{\link{wwwhr}},
\code{\link{wwwmin}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
