% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{ushopefl}
\alias{ushopefl}
\title{Thinking us how often feel hopeful?}
\source{
General Social Survey https://gss.norc.org
}
\description{
ushopefl
}
\details{
Question Now I want to ask some questions about how you feel when you think about this country, the United States. When you think about the United States:
F. When you think about the United States, how often do you feel HOPEFUL? Do you feel hopeful:
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} most of the time
\item \verb{[3]} some of the time
\item \verb{[4]} rarely
\item \verb{[5]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1794/vshow}{GSS Data Explorer page for \code{ushopefl}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab always \tab most of the time \tab some of the time \tab rarely \tab never \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 2199 \tab 148 \tab 238 \tab 241 \tab 50 \tab 15 \tab 11 \tab 2 \tab - \tab 2904 \cr
   Total \tab 2199 \tab 148 \tab 238 \tab 241 \tab 50 \tab 15 \tab 11 \tab 2 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{angrlast}},
\code{\link{angry}},
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{anxious}},
\code{\link{ashamed}},
\code{\link{atease}},
\code{\link{badbrks}},
\code{\link{beplesnt}},
\code{\link{calm}},
\code{\link{chnang1}},
\code{\link{chnang10}},
\code{\link{chnang11}},
\code{\link{chnang12}},
\code{\link{chnang13}},
\code{\link{chnang14}},
\code{\link{chnang15}},
\code{\link{chnang16}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{chnang6}},
\code{\link{chnang7}},
\code{\link{chnang8}},
\code{\link{chnang9}},
\code{\link{contentd}},
\code{\link{embarrss}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{hapfeel}},
\code{\link{howangry}},
\code{\link{likeoth}},
\code{\link{litcntrl}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{madat1}},
\code{\link{madat10}},
\code{\link{madat11}},
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat19}},
\code{\link{madat2}},
\code{\link{madat20}},
\code{\link{madat21}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat5}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}},
\code{\link{mostluck}},
\code{\link{noemote}},
\code{\link{noplan}},
\code{\link{notam}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{othresp}},
\code{\link{outraged}},
\code{\link{ovrjoyed}},
\code{\link{proud}},
\code{\link{reactok}},
\code{\link{restless}},
\code{\link{sad}},
\code{\link{selfresp}},
\code{\link{shakeblu}},
\code{\link{showangr}},
\code{\link{showfeel}},
\code{\link{thnkangr}},
\code{\link{usangry}},
\code{\link{usenthus}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usupset}},
\code{\link{usworry}},
\code{\link{whnangry}},
\code{\link{worried}}

Other United States: 
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amimp}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amprogrn}},
\code{\link{amproud}},
\code{\link{amproud1}},
\code{\link{amrank}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{askcomp}},
\code{\link{askcrime}},
\code{\link{askdrink}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{askmentl}},
\code{\link{asksexor}},
\code{\link{asnchng}},
\code{\link{belikeus}},
\code{\link{blkchng}},
\code{\link{bugging}},
\code{\link{camwrk}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{citizen}},
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{compfin}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{corrupt}},
\code{\link{decsorgs}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{demworks}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{engoffcl}},
\code{\link{excldimm}},
\code{\link{finanqs}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{freetrde}},
\code{\link{frterror}},
\code{\link{futenrgy}},
\code{\link{guerilla}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}},
\code{\link{hispchng}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immcrime}},
\code{\link{immcult}},
\code{\link{immfare}},
\code{\link{immideas}},
\code{\link{immjobs}},
\code{\link{imports}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intlincs}},
\code{\link{jewschng}},
\code{\link{landwar}},
\code{\link{leakinfo}},
\code{\link{lessnuke}},
\code{\link{lessprd}},
\code{\link{letin}},
\code{\link{letin1}},
\code{\link{lietest}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{morenuke}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{nonuke}},
\code{\link{notam}},
\code{\link{nuclrwar}},
\code{\link{nukewar}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{punarms}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}},
\code{\link{refugees}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{rptcowrk}},
\code{\link{satdemoc}},
\code{\link{secdocs}},
\code{\link{secprvcy}},
\code{\link{sectech}},
\code{\link{shortcom}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{srchwrk}},
\code{\link{stoprndm}},
\code{\link{takearms}},
\code{\link{taketrck}},
\code{\link{taphme}},
\code{\link{tapphone}},
\code{\link{tapwrk}},
\code{\link{testdrug}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{usamind}},
\code{\link{usangry}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{usdoenuf}},
\code{\link{usenthus}},
\code{\link{usfrustr}},
\code{\link{ushisp}},
\code{\link{usintl}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{ussat}},
\code{\link{usspy}},
\code{\link{usterror}},
\code{\link{usun}},
\code{\link{usupset}},
\code{\link{uswar}},
\code{\link{uswary}},
\code{\link{uswht}},
\code{\link{usworry}},
\code{\link{whtchng}},
\code{\link{wotrial}},
\code{\link{wrldgovt}}
}
\concept{Emotions}
\concept{United States}
\keyword{variable}
