% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{usmca1}
\alias{usmca1}
\title{R's familiarity with usmca}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmca1
}
\details{
Question How much have you heard or read about the United States-Mexico-Canada Agreement (USMCA)?
}
\section{Values}{

\itemize{
\item \verb{[1]} a lot
\item \verb{[2]} quite a bit
\item \verb{[3]} not much
\item \verb{[4]} nothing at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8185/vshow}{GSS Data Explorer page for \code{usmca1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab a lot \tab quite a bit \tab not much \tab nothing at all \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 38 \tab 124 \tab 677 \tab 721 \tab 62 \tab 1669 \tab 6 \tab 12 \tab 3309 \cr
   Total \tab 0 \tab 38 \tab 124 \tab 677 \tab 721 \tab 62 \tab 1669 \tab 6 \tab 12 \tab 3309 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{usmca2}}
}
\concept{ISSP Citizenship}
\keyword{variable}
