% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{usmca2}
\alias{usmca2}
\title{Does the us benefit from usmca}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmca2
}
\details{
Question Generally speaking, would you say that America benefits or does not benefit from being a member of United States-Mexico-Canada Agreement (USMCA)?
}
\section{Values}{

\itemize{
\item \verb{[1]} greatly benefits
\item \verb{[2]} largely benefits
\item \verb{[3]} somewhat benefits
\item \verb{[4]} benefits only a little
\item \verb{[5]} does not benefit at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8186/vshow}{GSS Data Explorer page for \code{usmca2}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab greatly benefits \tab largely benefits \tab somewhat benefits \tab benefits only a little \tab does not benefit at all \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 53 \tab 146 \tab 269 \tab 71 \tab 37 \tab 321 \tab 2397 \tab 1 \tab 14 \tab 3309 \cr
   Total \tab 0 \tab 53 \tab 146 \tab 269 \tab 71 \tab 37 \tab 321 \tab 2397 \tab 1 \tab 14 \tab 3309 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{usmca1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
