% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_256.R
\name{usmixed}
\alias{usmixed}
\title{What percentage of population is of mixed heritage}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmixed
}
\details{
Question People may come from one racial group such as being White, Black/African-American, Hispanic, Asian, or American Indian, or people may come from two or more of these groups. What percent of people in the United States today do you think have parents or their ancestors from two or more racial groups?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2316/vshow}{GSS Data Explorer page for \code{usmixed}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 12 \tab 13 \tab 15 \tab 16 \tab 17 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 27 \tab 28 \tab 3 \tab 30 \tab 32 \tab 33 \tab 35 \tab 38 \tab 4 \tab 40 \tab 42 \tab 45 \tab 47 \tab 5 \tab 50 \tab 51 \tab 55 \tab 6 \tab 60 \tab 62 \tab 63 \tab 65 \tab 68 \tab 7 \tab 70 \tab 72 \tab 74 \tab 75 \tab 8 \tab 80 \tab 85 \tab 88 \tab 90 \tab 92 \tab 95 \tab 96 \tab 97 \tab 98 \tab 99 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 3 \tab 7 \tab 73 \tab 36 \tab 2 \tab 1 \tab 70 \tab 1 \tab 2 \tab 2 \tab 6 \tab 123 \tab 1 \tab 1 \tab 2 \tab 1 \tab 85 \tab 3 \tab 3 \tab 6 \tab 128 \tab 1 \tab 2 \tab 30 \tab 1 \tab 3 \tab 118 \tab 1 \tab 19 \tab 1 \tab 34 \tab 145 \tab 1 \tab 4 \tab 3 \tab 68 \tab 1 \tab 1 \tab 14 \tab 2 \tab 5 \tab 43 \tab 1 \tab 1 \tab 62 \tab 3 \tab 63 \tab 20 \tab 1 \tab 43 \tab 1 \tab 14 \tab 1 \tab 1 \tab 9 \tab 3 \tab 81 \tab 41 \tab 2817 \cr
   Total \tab 1419 \tab 3 \tab 7 \tab 73 \tab 36 \tab 2 \tab 1 \tab 70 \tab 1 \tab 2 \tab 2 \tab 6 \tab 123 \tab 1 \tab 1 \tab 2 \tab 1 \tab 85 \tab 3 \tab 3 \tab 6 \tab 128 \tab 1 \tab 2 \tab 30 \tab 1 \tab 3 \tab 118 \tab 1 \tab 19 \tab 1 \tab 34 \tab 145 \tab 1 \tab 4 \tab 3 \tab 68 \tab 1 \tab 1 \tab 14 \tab 2 \tab 5 \tab 43 \tab 1 \tab 1 \tab 62 \tab 3 \tab 63 \tab 20 \tab 1 \tab 43 \tab 1 \tab 14 \tab 1 \tab 1 \tab 9 \tab 3 \tab 81 \tab 41 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkchng}},
\code{\link{hispchng}},
\code{\link{jewschng}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{whtchng}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
