% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{usrep}
\alias{usrep}
\title{Did r correctly name us representative}
\source{
General Social Survey https://gss.norc.org
}
\description{
usrep
}
\details{
Question What about the Congressman from this district? Do you happen to know (his/her) name?
}
\section{Values}{

\itemize{
\item \verb{[1]} correct
\item \verb{[2]} incorrect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1036/vshow}{GSS Data Explorer page for \code{usrep}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab correct \tab incorrect \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 621 \tab 158 \tab 1002 \tab 38 \tab 1819 \cr
   Total \tab 0 \tab 621 \tab 158 \tab 1002 \tab 38 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}}
}
\concept{Socio-political participation}
\keyword{variable}
