% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{usspy}
\alias{usspy}
\title{U.s. spies foreign govts more threat 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
usspy
}
\details{
Question Are the following threats to the United States greater, about the same, or less today than they were 10 years ago?
A. Spying by U.S. citizens for foreign countries
}
\section{Values}{

\itemize{
\item \code{1} greater
\item \code{2} about the same
\item \code{3} less
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5257/vshow}{GSS Data Explorer page for \code{usspy}}.
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askcrime}},
\code{\link{askdrink}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{askmentl}},
\code{\link{asksexor}},
\code{\link{bugging}},
\code{\link{camwrk}},
\code{\link{chkfinan}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{chktravl}},
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{compfin}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{lietest}},
\code{\link{nuclrwar}},
\code{\link{punarms}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}},
\code{\link{rptcowrk}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secdocs}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{sectech}},
\code{\link{secterr}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{srchwrk}},
\code{\link{takearms}},
\code{\link{taketrck}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{testdrug}},
\code{\link{usterror}}
}
\concept{National Security}
\keyword{variable}
