% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{usualhrs}
\alias{usualhrs}
\title{How many hrs/week do you usually work}
\source{
General Social Survey https://gss.norc.org
}
\description{
usualhrs
}
\details{
Question How many hours a week do you usually work, at all paid jobs?
HOURS:
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6292/vshow}{GSS Data Explorer page for \code{usualhrs}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 27 \tab 28 \tab 29 \tab 30 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 48 \tab 5 \tab 50 \tab 52 \tab 55 \tab 58 \tab 6 \tab 60 \tab 65 \tab 67 \tab 7 \tab 70 \tab 72 \tab 75 \tab 8 \tab 80 \tab 99 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2273 \tab 1 \tab 8 \tab 6 \tab 7 \tab 4 \tab 1 \tab 18 \tab 2 \tab 2 \tab 2 \tab 3 \tab 15 \tab 3 \tab 2 \tab 1 \tab 24 \tab 10 \tab 2 \tab 2 \tab 22 \tab 5 \tab 5 \tab 10 \tab 4 \tab 2 \tab 212 \tab 4 \tab 2 \tab 6 \tab 4 \tab 39 \tab 1 \tab 8 \tab 1 \tab 53 \tab 6 \tab 10 \tab 1 \tab 2 \tab 32 \tab 11 \tab 1 \tab 1 \tab 8 \tab 1 \tab 2 \tab 1 \tab 7 \tab 1 \tab 4 \tab 15 \tab 2867 \cr
   Total \tab 2273 \tab 1 \tab 8 \tab 6 \tab 7 \tab 4 \tab 1 \tab 18 \tab 2 \tab 2 \tab 2 \tab 3 \tab 15 \tab 3 \tab 2 \tab 1 \tab 24 \tab 10 \tab 2 \tab 2 \tab 22 \tab 5 \tab 5 \tab 10 \tab 4 \tab 2 \tab 212 \tab 4 \tab 2 \tab 6 \tab 4 \tab 39 \tab 1 \tab 8 \tab 1 \tab 53 \tab 6 \tab 10 \tab 1 \tab 2 \tab 32 \tab 11 \tab 1 \tab 1 \tab 8 \tab 1 \tab 2 \tab 1 \tab 7 \tab 1 \tab 4 \tab 15 \tab 2867 \cr
}
}

\seealso{
Other Flexible Work: 
\code{\link{mosthrs}},
\code{\link{paidhow}}
}
\concept{Flexible Work}
\keyword{variable}
