% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{uswary}
\alias{uswary}
\title{Expect u.s. in world war in 10 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
uswary
}
\details{
Question Do you expect the United States to fight in another world war within the next ten years?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1976, 1985, 1986 \cr
   A/- \tab 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5485/vshow}{GSS Data Explorer page for \code{uswary}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1976 \tab 778 \tab 313 \tab 360 \tab 45 \tab 3 \tab - \tab 1499 \cr
   1985 \tab 751 \tab 334 \tab 397 \tab 47 \tab 5 \tab - \tab 1534 \cr
   1986 \tab - \tab 663 \tab 733 \tab 47 \tab 27 \tab - \tab 1470 \cr
   1988 \tab 493 \tab 394 \tab 555 \tab 38 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 315 \tab 637 \tab 51 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 255 \tab 635 \tab 37 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 465 \tab 505 \tab 54 \tab - \tab - \tab 1517 \cr
   1993 \tab 526 \tab 505 \tab 536 \tab 38 \tab 1 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 822 \tab 1007 \tab 90 \tab 58 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 756 \tab 1085 \tab 88 \tab 31 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 815 \tab 836 \tab 86 \tab 134 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 668 \tab 1086 \tab 115 \tab 8 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 575 \tab 269 \tab 64 \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 460 \tab 368 \tab 78 \tab - \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 1124 \tab 732 \tab 136 \tab - \tab - \tab 4510 \cr
   2008 \tab 694 \tab 652 \tab 590 \tab 22 \tab 65 \tab - \tab 2023 \cr
   2010 \tab 1377 \tab 336 \tab 302 \tab 29 \tab - \tab - \tab 2044 \cr
   2012 \tab 1338 \tab 307 \tab 292 \tab 37 \tab - \tab - \tab 1974 \cr
   2014 \tab 1687 \tab 453 \tab 342 \tab 56 \tab - \tab - \tab 2538 \cr
   2016 \tab 1956 \tab 513 \tab 352 \tab 46 \tab - \tab - \tab 2867 \cr
   2018 \tab 1563 \tab 378 \tab 350 \tab 56 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 2936 \tab 334 \tab 218 \tab 56 \tab - \tab - \tab 3544 \cr
   Total \tab 25708 \tab 11437 \tab 12187 \tab 1316 \tab 338 \tab 4032 \tab 55018 \cr
}
}

\seealso{
Other War and International: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{commun}},
\code{\link{commun10}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{israel}},
\code{\link{japan}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}}

Other Foreign Affairs: 
\code{\link{intintl}},
\code{\link{knwforgn}},
\code{\link{landwar}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{moretrde}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}}

Other Split Ballots: 
\code{\link{abany}},
\code{\link{abcare}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abdefect}},
\code{\link{abfirm}},
\code{\link{abhlth}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{aidsads}},
\code{\link{aidsfare}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{alienat6}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{blkcol}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{class}()},
\code{\link{classy}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{colaff}},
\code{\link{colop}},
\code{\link{conarmy}},
\code{\link{conarmyy}},
\code{\link{conbus}},
\code{\link{conbusy}},
\code{\link{conclerg}},
\code{\link{conclery}},
\code{\link{coneduc}},
\code{\link{coneducy}},
\code{\link{confed}},
\code{\link{confedy}},
\code{\link{confinan}},
\code{\link{confinay}},
\code{\link{conjudge}},
\code{\link{conjudgy}},
\code{\link{conlabor}},
\code{\link{conlaboy}},
\code{\link{conlegis}},
\code{\link{conlegiy}},
\code{\link{conmedic}},
\code{\link{conmediy}},
\code{\link{conpress}},
\code{\link{conpresy}},
\code{\link{consci}},
\code{\link{consciy}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{contv}},
\code{\link{contvy}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{creator}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{crtrheal}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{draft}},
\code{\link{draftco}},
\code{\link{draftcol}},
\code{\link{draftdef}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{father}},
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fegive}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fejoin}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{feserve}},
\code{\link{fethink}},
\code{\link{fewrite}},
\code{\link{forbdcom}},
\code{\link{forbdmar}},
\code{\link{forbdrac}},
\code{\link{form}},
\code{\link{formwt}},
\code{\link{friend}},
\code{\link{frndking}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{hapmar}},
\code{\link{happy}},
\code{\link{happy7}},
\code{\link{healer}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{israel}},
\code{\link{japan}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobpeop}},
\code{\link{jobpromo}},
\code{\link{jobresp}},
\code{\link{jobrise}},
\code{\link{jobsafe}},
\code{\link{jobsec}},
\code{\link{jobsoc}},
\code{\link{judge}},
\code{\link{judgeluv}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{king}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{liberatr}},
\code{\link{lover}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mapa}},
\code{\link{master}},
\code{\link{mastersp}},
\code{\link{meserve}},
\code{\link{mother}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{opfamily}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{partners}},
\code{\link{partners_8822}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racimp}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{rank}()},
\code{\link{rdeemlib}},
\code{\link{redeemer}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{russia}},
\code{\link{satcity}},
\code{\link{satfam}},
\code{\link{satfrnd}},
\code{\link{sathealt}},
\code{\link{sathobby}},
\code{\link{sexfreq}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{spouse}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{tax}},
\code{\link{taxserve}},
\code{\link{trust}},
\code{\link{trusty}},
\code{\link{uswar}},
\code{\link{xmovie1}}

Other United States: 
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amimp}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amprogrn}},
\code{\link{amproud}},
\code{\link{amproud1}},
\code{\link{amrank}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{askcomp}},
\code{\link{askcrime}},
\code{\link{askdrink}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{askmentl}},
\code{\link{asksexor}},
\code{\link{asnchng}},
\code{\link{belikeus}},
\code{\link{blkchng}},
\code{\link{bugging}},
\code{\link{camwrk}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{citizen}},
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{compfin}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{corrupt}},
\code{\link{decsorgs}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{demworks}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{engoffcl}},
\code{\link{excldimm}},
\code{\link{finanqs}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{freetrde}},
\code{\link{frterror}},
\code{\link{futenrgy}},
\code{\link{guerilla}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}},
\code{\link{hispchng}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immcrime}},
\code{\link{immcult}},
\code{\link{immfare}},
\code{\link{immideas}},
\code{\link{immjobs}},
\code{\link{imports}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intlincs}},
\code{\link{jewschng}},
\code{\link{landwar}},
\code{\link{leakinfo}},
\code{\link{lessnuke}},
\code{\link{lessprd}},
\code{\link{letin}},
\code{\link{letin1}},
\code{\link{lietest}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{morenuke}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{nataid}},
\code{\link{nataidy}},
\code{\link{nataidz}},
\code{\link{natarms}},
\code{\link{natarmsy}},
\code{\link{natarmsz}},
\code{\link{natchld}},
\code{\link{natcity}},
\code{\link{natcityy}},
\code{\link{natcityz}},
\code{\link{natcrime}},
\code{\link{natcrimy}},
\code{\link{natcrimz}},
\code{\link{natdrug}},
\code{\link{natdrugy}},
\code{\link{nateduc}},
\code{\link{nateducy}},
\code{\link{nateducz}},
\code{\link{natenrgy}},
\code{\link{natenvir}},
\code{\link{natenviy}},
\code{\link{natenviz}},
\code{\link{natfare}},
\code{\link{natfarey}},
\code{\link{natfarez}},
\code{\link{natheal}},
\code{\link{nathealy}},
\code{\link{nathealz}},
\code{\link{natmass}},
\code{\link{natmassz}},
\code{\link{natpark}},
\code{\link{natparkz}},
\code{\link{natrace}},
\code{\link{natracey}},
\code{\link{natracez}},
\code{\link{natroad}},
\code{\link{natroadz}},
\code{\link{natsci}},
\code{\link{natsoc}},
\code{\link{natsocz}},
\code{\link{natspac}},
\code{\link{natspacy}},
\code{\link{natspacz}},
\code{\link{nonuke}},
\code{\link{notam}},
\code{\link{nuclrwar}},
\code{\link{nukewar}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{punarms}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}},
\code{\link{refugees}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}},
\code{\link{rptcowrk}},
\code{\link{satdemoc}},
\code{\link{secdocs}},
\code{\link{secprvcy}},
\code{\link{sectech}},
\code{\link{shortcom}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{srchwrk}},
\code{\link{stoprndm}},
\code{\link{takearms}},
\code{\link{taketrck}},
\code{\link{taphme}},
\code{\link{tapphone}},
\code{\link{tapwrk}},
\code{\link{testdrug}},
\code{\link{topprob1}},
\code{\link{topprob1_next}},
\code{\link{topprob2}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}},
\code{\link{usamind}},
\code{\link{usangry}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{usdoenuf}},
\code{\link{usenthus}},
\code{\link{usfrustr}},
\code{\link{ushisp}},
\code{\link{ushopefl}},
\code{\link{usintl}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{ussat}},
\code{\link{usspy}},
\code{\link{usterror}},
\code{\link{usun}},
\code{\link{usupset}},
\code{\link{uswar}},
\code{\link{uswht}},
\code{\link{usworry}},
\code{\link{whtchng}},
\code{\link{wotrial}},
\code{\link{wrldgovt}}
}
\concept{Foreign Affairs}
\concept{Split Ballots}
\concept{United States}
\concept{War and International}
\keyword{variable}
