% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{uswarynv}
\alias{uswarynv}
\title{Expect u.s. in war within 10 years (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
uswarynv
}
\details{
Question Do you expect the United States to fight in another world war within the next ten years?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} undecided (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7604/vshow}{GSS Data Explorer page for \code{uswarynv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 3358 \tab 268 \tab 395 \tab 7 \tab 4 \tab 4032 \cr
   2022 \tab 3259 \tab 151 \tab 134 \tab - \tab - \tab 3544 \cr
   Total \tab 6617 \tab 419 \tab 529 \tab 7 \tab 4 \tab 7576 \cr
}
}

\keyword{variable}
