% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{uswaryv}
\alias{uswaryv}
\title{Expect u.s. in war within 10 years (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
uswaryv
}
\details{
Question Do you expect the United States to fight in another world war within the next ten years?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} undecided (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7584/vshow}{GSS Data Explorer page for \code{uswaryv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 3346 \tab 213 \tab 363 \tab 109 \tab 1 \tab 4032 \cr
   2022 \tab 3264 \tab 136 \tab 92 \tab 51 \tab 1 \tab 3544 \cr
   Total \tab 6610 \tab 349 \tab 455 \tab 160 \tab 2 \tab 7576 \cr
}
}

\keyword{variable}
