% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{uswht}
\alias{uswht}
\title{What percentage of population is white}
\source{
General Social Survey https://gss.norc.org
}
\description{
uswht
}
\details{
Question Just your best guess–what percentage of the United States population is each group? (INSTRUCTION: PERCENTAGES DO NOT HAVE TO ADD TO 100pct AND THE LISTED GROUPS MAY OVERLAP. ENCOURAGE ESTIMATES FOR ALL GROUPS, BUT DISCOURAGE RESPONDENTS REVISING THEIR ESTIMATES.)
A. Whites?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2310/vshow}{GSS Data Explorer page for \code{uswht}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 15 \tab 20 \tab 22 \tab 23 \tab 25 \tab 28 \tab 30 \tab 33 \tab 35 \tab 37 \tab 38 \tab 40 \tab 41 \tab 42 \tab 43 \tab 45 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 55 \tab 56 \tab 57 \tab 58 \tab 6 \tab 60 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 7 \tab 70 \tab 72 \tab 74 \tab 75 \tab 79 \tab 8 \tab 80 \tab 82 \tab 85 \tab 89 \tab 90 \tab 95 \tab 98 \tab 99 \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 1 \tab 2 \tab 10 \tab 1 \tab 9 \tab 1 \tab 1 \tab 15 \tab 1 \tab 33 \tab 4 \tab 18 \tab 1 \tab 3 \tab 103 \tab 2 \tab 1 \tab 2 \tab 25 \tab 2 \tab 4 \tab 3 \tab 1 \tab 222 \tab 4 \tab 6 \tab 1 \tab 33 \tab 1 \tab 3 \tab 3 \tab 2 \tab 303 \tab 4 \tab 5 \tab 2 \tab 84 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 165 \tab 3 \tab 1 \tab 109 \tab 1 \tab 1 \tab 80 \tab 1 \tab 10 \tab 1 \tab 14 \tab 4 \tab 1 \tab 1 \tab 57 \tab 23 \tab 2817 \cr
   Total \tab 1419 \tab 1 \tab 2 \tab 10 \tab 1 \tab 9 \tab 1 \tab 1 \tab 15 \tab 1 \tab 33 \tab 4 \tab 18 \tab 1 \tab 3 \tab 103 \tab 2 \tab 1 \tab 2 \tab 25 \tab 2 \tab 4 \tab 3 \tab 1 \tab 222 \tab 4 \tab 6 \tab 1 \tab 33 \tab 1 \tab 3 \tab 3 \tab 2 \tab 303 \tab 4 \tab 5 \tab 2 \tab 84 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 165 \tab 3 \tab 1 \tab 109 \tab 1 \tab 1 \tab 80 \tab 1 \tab 10 \tab 1 \tab 14 \tab 4 \tab 1 \tab 1 \tab 57 \tab 23 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{othhome}},
\code{\link{spklangw}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
