% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_207.R
\name{valeduc}
\alias{valeduc}
\title{Estimated value contributed in education}
\source{
General Social Survey https://gss.norc.org
}
\description{
valeduc
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
B. Education
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1961/vshow}{GSS Data Explorer page for \code{valeduc}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 12 \tab 15 \tab 1500 \tab 20 \tab 200 \tab 25 \tab 30 \tab 300 \tab 3000 \tab 5 \tab 50 \tab 500 \tab 75 \tab 80 \tab 99996 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2685 \tab 93 \tab 3 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 10 \tab 75 \tab 2904 \cr
   Total \tab 2685 \tab 93 \tab 3 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 10 \tab 75 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{totart}},
\code{\link{totfound}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totwork}},
\code{\link{totyouth}},
\code{\link{valhlth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
