% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_208.R
\name{valfound}
\alias{valfound}
\title{Estimated value contributed in priv & comm. found}
\source{
General Social Survey https://gss.norc.org
}
\description{
valfound
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
L. Private and community foundations
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1971/vshow}{GSS Data Explorer page for \code{valfound}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 150 \tab 18 \tab 20 \tab 200 \tab 2000 \tab 210 \tab 25 \tab 250 \tab 28 \tab 30 \tab 300 \tab 40 \tab 50 \tab 500 \tab 65 \tab 70 \tab 90 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2700 \tab 82 \tab 1 \tab 7 \tab 1 \tab 2 \tab 1 \tab 6 \tab 9 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 9 \tab 58 \tab 2904 \cr
   Total \tab 2700 \tab 82 \tab 1 \tab 7 \tab 1 \tab 2 \tab 1 \tab 6 \tab 9 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 9 \tab 58 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{valart}},
\code{\link{valenvir}},
\code{\link{valhuman}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
