% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{valgiven}
\alias{valgiven}
\title{Total donations past year r and immediate family}
\source{
General Social Survey https://gss.norc.org
}
\description{
valgiven
}
\details{
Question Altogether, what was the total dollar value of all donations you and your immediate family made in the past year towards religious and charitable purposes?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/C/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2895/vshow}{GSS Data Explorer page for \code{valgiven}}.
}

\seealso{
Other Altruism: 
\code{\link{volmonth}}

Other Altruism: 
\code{\link{volmonth}}
}
\concept{Altruism}
\keyword{variable}
