% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_208.R
\name{valhuman}
\alias{valhuman}
\title{Estimated value contributed in human services}
\source{
General Social Survey https://gss.norc.org
}
\description{
valhuman
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
D. Human services
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1963/vshow}{GSS Data Explorer page for \code{valhuman}}.

Counts by year:\tabular{lllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 1200 \tab 15 \tab 150 \tab 1500 \tab 20 \tab 200 \tab 2000 \tab 25 \tab 250 \tab 30 \tab 300 \tab 400 \tab 5 \tab 50 \tab 500 \tab 60 \tab 600 \tab 75 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2674 \tab 94 \tab 5 \tab 11 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 4 \tab 1 \tab 6 \tab 8 \tab 1 \tab 1 \tab 2 \tab 8 \tab 60 \tab 2904 \cr
   Total \tab 2674 \tab 94 \tab 5 \tab 11 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 4 \tab 1 \tab 6 \tab 8 \tab 1 \tab 1 \tab 2 \tab 8 \tab 60 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{valart}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
