% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_209.R
\name{valoth}
\alias{valoth}
\title{Estimated value contributed in other (specify)}
\source{
General Social Survey https://gss.norc.org
}
\description{
valoth
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
O. Other (SPECIFY)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1974/vshow}{GSS Data Explorer page for \code{valoth}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 120 \tab 14000 \tab 15 \tab 150 \tab 20 \tab 200 \tab 25 \tab 300 \tab 35 \tab 45 \tab 5 \tab 50 \tab 500 \tab 65 \tab 75 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2807 \tab 39 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 3 \tab 24 \tab 2904 \cr
   Total \tab 2807 \tab 39 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 3 \tab 24 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{homeless}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{valinfrm}},
\code{\link{valintl}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
