% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_208.R
\name{valpol}
\alias{valpol}
\title{Estimated value contributed in political org. etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
valpol
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
J. Political organizations or campaigns
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1969/vshow}{GSS Data Explorer page for \code{valpol}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 100 \tab 25 \tab 50 \tab 500 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2812 \tab 35 \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab 43 \tab 2904 \cr
   Total \tab 2812 \tab 35 \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab 43 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{valart}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhuman}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
