% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_208.R
\name{valrelig}
\alias{valrelig}
\title{Estimated value contributed in religious org}
\source{
General Social Survey https://gss.norc.org
}
\description{
valrelig
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
C. Religious organizations
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1962/vshow}{GSS Data Explorer page for \code{valrelig}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 1200 \tab 15 \tab 150 \tab 20 \tab 200 \tab 2000 \tab 25 \tab 250 \tab 2500 \tab 2800 \tab 30 \tab 300 \tab 3000 \tab 350 \tab 3500 \tab 40 \tab 400 \tab 4000 \tab 5 \tab 50 \tab 500 \tab 600 \tab 6000 \tab 625 \tab 70 \tab 700 \tab 75 \tab 750 \tab 80 \tab 800 \tab 9000 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2274 \tab 290 \tab 2 \tab 24 \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab 11 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 3 \tab 6 \tab 2 \tab 1 \tab 1 \tab 5 \tab 7 \tab 1 \tab 2 \tab 5 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 33 \tab 196 \tab 2904 \cr
   Total \tab 2274 \tab 290 \tab 2 \tab 24 \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab 11 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 3 \tab 6 \tab 2 \tab 1 \tab 1 \tab 5 \tab 7 \tab 1 \tab 2 \tab 5 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 33 \tab 196 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{valart}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhuman}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valwork}},
\code{\link{valyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
