% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_208.R
\name{valyouth}
\alias{valyouth}
\title{Estimated value contributed in youth development}
\source{
General Social Survey https://gss.norc.org
}
\description{
valyouth
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
K. Youth development
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1970/vshow}{GSS Data Explorer page for \code{valyouth}}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 125 \tab 1250 \tab 15 \tab 175 \tab 20 \tab 200 \tab 2222 \tab 25 \tab 30 \tab 300 \tab 40 \tab 400 \tab 50 \tab 600 \tab 75 \tab 80 \tab 800 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2706 \tab 94 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 65 \tab 2904 \cr
   Total \tab 2706 \tab 94 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 65 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{valart}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhuman}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
