% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{vaxsafe}
\alias{vaxsafe}
\title{Vaccines are safe}
\source{
General Social Survey https://gss.norc.org
}
\description{
vaxsafe
}
\details{
Question (Please tell us the extent to which you agree or disagree with the following statements:) Vaccines are safe.

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7950/vshow}{GSS Data Explorer page for \code{vaxsafe}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 407 \tab 430 \tab 332 \tab 43 \tab 20 \tab 3 \tab 3544 \cr
   Total \tab 2309 \tab 407 \tab 430 \tab 332 \tab 43 \tab 20 \tab 3 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{fluvax}},
\code{\link{malekids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}}
}
\concept{GSS Next}
\keyword{variable}
