% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{verdict}
\alias{verdict}
\title{Worse type of judicial mistake}
\source{
General Social Survey https://gss.norc.org
}
\description{
verdict
}
\details{
Question All systems of justice make mistakes, but which do you think is worse?
To convict an innocent person
or
To let a guilty person go free
}
\section{Values}{

\itemize{
\item \code{1} convict an innocent person
\item \code{2} let a guilty person go free
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4002/vshow}{GSS Data Explorer page for \code{verdict}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{databank}},
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{progtax}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
