% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{version}
\alias{version}
\title{Version of questionnaire}
\source{
General Social Survey https://gss.norc.org
}
\description{
version
}
\details{
Question Version Administered.
}
\section{Values}{

\itemize{
\item \code{1} version 1
\item \code{2} version 2
\item \code{3} version 3
\item \code{4} version 4
\item \code{5} version 5
\item \code{6} version 6
\item \code{7} version 7
\item \code{8} version 8
\item \code{9} version 9
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5575/vshow}{GSS Data Explorer page for \code{version}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{sampcode}},
\code{\link{sample}()}

Other Sample: 
\code{\link{ballot}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{sampcode}},
\code{\link{sample}()}
}
\concept{Sample}
\concept{Sampling and Weights}
\keyword{variable}
