% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{vetaid}
\alias{vetaid}
\title{Any in hh receive mil or vet benefits}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetaid
}
\details{
Question Are you or any members of your present household currently receiving any pay or benefits from either the military or the Veterans Administration?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/828/vshow}{GSS Data Explorer page for \code{vetaid}}.
}

\seealso{
Other Core: 
\code{\link{defwrkev}},
\code{\link{guerilla}},
\code{\link{jobtrain}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()}
}
\concept{Core}
\concept{Veterans' Aid}
\keyword{variable}
