% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{vetfam}
\alias{vetfam}
\title{Family members served in armed forces?}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetfam
}
\details{
Question Have any other members of your immediate family--that is, your (spouse/parent/children/brothers or sisters)--ever served in the armed forces?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/812/vshow}{GSS Data Explorer page for \code{vetfam}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1982 \tab - \tab 1297 \tab 551 \tab 12 \tab 1860 \cr
   1984 \tab - \tab 1066 \tab 381 \tab 26 \tab 1473 \cr
   Total \tab 0 \tab 2363 \tab 932 \tab 38 \tab 3333 \cr
}
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfamnw}}
}
\concept{Core}
\keyword{variable}
