% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{vetfamnw}
\alias{vetfamnw}
\title{Any family members in armed forces now}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetfamnw
}
\details{
Question Have any other members of your immediate family--that is, your (spouse/parent/children/brothers or sisters)--ever served in the armed forces?
A. Are any other members of your immediate family serving in the Armed Forces now?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/813/vshow}{GSS Data Explorer page for \code{vetfamnw}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1984 \tab 381 \tab 144 \tab 912 \tab 36 \tab 1473 \cr
   Total \tab 381 \tab 144 \tab 912 \tab 36 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfam}}
}
\concept{Core}
\keyword{variable}
