% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{vetkind}
\alias{vetkind}
\title{Branch of service}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetkind
}
\details{
Question Have you ever been on active duty for military training or service for two consecutive months or more?
IF YES: What was your total time on active duty?
A. In what branch of the service was that?
}
\section{Values}{

\itemize{
\item \code{0} several
\item \code{1} air force guard
\item \code{2} air force
\item \code{3} navy
\item \code{4} army
\item \code{5} national guard
\item \code{6} marines
\item \code{7} coast guard
\item \code{8} public health
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab - \tab full \cr
   1975 \tab - \tab full \cr
   1977 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5286/vshow}{GSS Data Explorer page for \code{vetkind}}.
}

\seealso{
Other Military Service & Work: 
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{vetyears}},
\code{\link{workdy}},
\code{\link{workhr}}
}
\concept{Military Service & Work}
\keyword{variable}
