% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{vigfrnd}
\alias{vigfrnd}
\title{How willing r would make friends with x}
\source{
General Social Survey https://gss.norc.org
}
\description{
vigfrnd
}
\details{
Question How willing would you be (READ A-F) -- definitely willing, probably willing, probably unwilling, or definitely unwilling?
C. To make friends with NAME?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely willing
\item \verb{[2]} probably willing
\item \verb{[3]} probably unwilling
\item \verb{[4]} definitely unwilling
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2005/vshow}{GSS Data Explorer page for \code{vigfrnd}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 266 \tab 657 \tab 274 \tab 178 \tab 45 \tab 24 \tab 2904 \cr
   2006 \tab 3073 \tab 291 \tab 737 \tab 265 \tab 89 \tab 38 \tab 17 \tab 4510 \cr
   2018 \tab 1175 \tab 229 \tab 564 \tab 223 \tab 109 \tab 19 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 786 \tab 1958 \tab 762 \tab 376 \tab 102 \tab 70 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
