% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{viggrp}
\alias{viggrp}
\title{How willing r would have a group home for x}
\source{
General Social Survey https://gss.norc.org
}
\description{
viggrp
}
\details{
Question How willing would you be (READ A-F) -- definitely willing, probably willing, probably unwilling, or definitely unwilling?
E. To have a group home for people like NAME opened in your neighborhood?
}
\section{Values}{

\itemize{
\item \code{1} definitely willing
\item \code{2} probably willing
\item \code{3} probably unwilling
\item \code{4} definitely unwilling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2007/vshow}{GSS Data Explorer page for \code{viggrp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 263 \tab 572 \tab 272 \tab 237 \tab 74 \tab 26 \tab 2904 \cr
   2006 \tab 3073 \tab 270 \tab 601 \tab 309 \tab 187 \tab 53 \tab 17 \tab 4510 \cr
   2018 \tab 1175 \tab 275 \tab 491 \tab 227 \tab 124 \tab 25 \tab 31 \tab 2348 \cr
   Total \tab 5708 \tab 808 \tab 1664 \tab 808 \tab 548 \tab 152 \tab 74 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
