% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{viglabel}
\alias{viglabel}
\title{How likely x is experiencing vignettes}
\source{
General Social Survey https://gss.norc.org
}
\description{
viglabel
}
\details{
Question In your opinion, how likely is it that NAME is experiencing (READ A-E) -- very likely, somewhat likely, not very likely, or not at all likely? (IN PART E, READ ONLY THE PHRASE FOLLOWING THE LETTER OF THE VIGNETTE)
E. REFER TO VIGNETTE ID, CHECK BELOW AND READ ONLY 1 PHRASE OF THE FOLLOWING:
Vignette A: alcohol dependence
Vignette B: a major depression
Vignette C: schizophrenia
Vignette D: a drug problem
Vignette E: GO TO DECTREAT
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1998/vshow}{GSS Data Explorer page for \code{viglabel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 755 \tab 235 \tab 48 \tab 20 \tab 50 \tab 336 \tab 2904 \cr
   2018 \tab 1411 \tab 481 \tab 209 \tab 29 \tab 13 \tab 20 \tab 185 \tab 2348 \cr
   Total \tab 2871 \tab 1236 \tab 444 \tab 77 \tab 33 \tab 70 \tab 521 \tab 5252 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}}
}
\concept{Mental Health}
\keyword{variable}
