% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{vigmar}
\alias{vigmar}
\title{How willing r would have x marry in r's family}
\source{
General Social Survey https://gss.norc.org
}
\description{
vigmar
}
\details{
Question How willing would you be (READ A-F) -- definitely willing, probably willing, probably unwilling, or definitely unwilling?
F. To have NAME marry into your family?
}
\section{Values}{

\itemize{
\item \code{1} definitely willing
\item \code{2} probably willing
\item \code{3} probably unwilling
\item \code{4} definitely unwilling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2008/vshow}{GSS Data Explorer page for \code{vigmar}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 105 \tab 307 \tab 379 \tab 530 \tab 97 \tab 26 \tab 2904 \cr
   2006 \tab 3073 \tab 130 \tab 403 \tab 436 \tab 378 \tab 74 \tab 16 \tab 4510 \cr
   2018 \tab 1175 \tab 115 \tab 337 \tab 373 \tab 275 \tab 38 \tab 35 \tab 2348 \cr
   Total \tab 5708 \tab 350 \tab 1047 \tab 1188 \tab 1183 \tab 209 \tab 77 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
