% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{vignei}
\alias{vignei}
\title{How willing r would have x as a neighbor}
\source{
General Social Survey https://gss.norc.org
}
\description{
vignei
}
\details{
Question How willing would you be (READ A-F) -- definitely willing, probably willing, probably unwilling, or definitely unwilling?
A. To move next door to NAME?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely willing
\item \verb{[2]} probably willing
\item \verb{[3]} probably unwilling
\item \verb{[4]} definitely unwilling
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2003/vshow}{GSS Data Explorer page for \code{vignei}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 260 \tab 586 \tab 320 \tab 206 \tab 51 \tab 21 \tab 2904 \cr
   2006 \tab 3073 \tab 272 \tab 707 \tab 276 \tab 122 \tab 45 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 190 \tab 548 \tab 234 \tab 138 \tab 30 \tab 33 \tab 2348 \cr
   Total \tab 5708 \tab 722 \tab 1841 \tab 830 \tab 466 \tab 126 \tab 69 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
