% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_408.R
\name{vigvermy}
\alias{vigvermy}
\title{Vignette version}
\source{
General Social Survey https://gss.norc.org
}
\description{
vigvermy
}
\details{
Question First, I'm going to describe a person. Let's call (him/her) (NAME). After I read a description of (him/her) I'll ask you some questions about how you think and feel about
}
\section{Values}{

\itemize{
\item \code{1} vignette 1
\item \code{2} vignette 2
\item \code{3} vignette 3
\item \code{4} vignette 4
\item \code{5} vignette 5
\item \code{6} vignette 6
\item \code{7} vignette 7
\item \code{8} vignette 8
\item \code{9} vignette 9
\item \code{10} vignette 10
\item \code{11} vignette 11
\item \code{12} vignette 12
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3719/vshow}{GSS Data Explorer page for \code{vigvermy}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab vignette 1 \tab vignette 2 \tab vignette 3 \tab vignette 4 \tab vignette 5 \tab vignette 6 \tab vignette 7 \tab vignette 8 \tab vignette 9 \tab vignette 10 \tab vignette 11 \tab vignette 12 \tab Total \cr
   2006 \tab 2992 \tab 125 \tab 136 \tab 125 \tab 125 \tab 116 \tab 131 \tab 132 \tab 129 \tab 123 \tab 119 \tab 128 \tab 129 \tab 4510 \cr
   Total \tab 2992 \tab 125 \tab 136 \tab 125 \tab 125 \tab 116 \tab 131 \tab 132 \tab 129 \tab 123 \tab 119 \tab 128 \tab 129 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdodoc}},
\code{\link{mhdofam}},
\code{\link{mhdofrnd}},
\code{\link{mhdoheal}},
\code{\link{mhdomhp}},
\code{\link{mhdootc}},
\code{\link{mhdopsyc}},
\code{\link{mhdorel}},
\code{\link{mhproblm}}
}
\concept{Mental Health IV}
\keyword{variable}
