% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_128.R
\name{violasns}
\alias{violasns}
\title{Violent - not violent}
\source{
General Social Survey https://gss.norc.org
}
\description{
violasns
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
C. The next set asks if people in each group tend to be violence prone or if they tend not to be prone to violence?
4. Asian Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} violence-prone
\item \verb{[7]} not violence-prone
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1193/vshow}{GSS Data Explorer page for \code{violasns}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab violence-prone \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab not violence-prone \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 25 \tab 49 \tab 180 \tab 541 \tab 214 \tab 125 \tab 59 \tab 153 \tab 26 \tab 1372 \cr
   2000 \tab 1419 \tab 26 \tab 42 \tab 137 \tab 565 \tab 232 \tab 145 \tab 69 \tab 168 \tab 14 \tab 2817 \cr
   Total \tab 1419 \tab 51 \tab 91 \tab 317 \tab 1106 \tab 446 \tab 270 \tab 128 \tab 321 \tab 40 \tab 4189 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intljews}},
\code{\link{intlwhts}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
