% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_128.R
\name{violhsps}
\alias{violhsps}
\title{Violent - not violent}
\source{
General Social Survey https://gss.norc.org
}
\description{
violhsps
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
C. The next set asks if people in each group tend to be violence prone or if they tend not to be prone to violence?
5. Hispanic Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} violence-prone
\item \verb{[7]} not violence-prone
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1194/vshow}{GSS Data Explorer page for \code{violhsps}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab violence-prone \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab not violence-prone \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab - \tab 63 \tab 170 \tab 294 \tab 457 \tab 152 \tab 69 \tab 24 \tab 117 \tab 26 \tab - \tab 1372 \cr
   2000 \tab 1419 \tab 49 \tab 118 \tab 303 \tab 553 \tab 138 \tab 67 \tab 27 \tab 128 \tab 15 \tab - \tab 2817 \cr
   Total \tab 1419 \tab 112 \tab 288 \tab 597 \tab 1010 \tab 290 \tab 136 \tab 51 \tab 245 \tab 41 \tab 0 \tab 4189 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Hispanics: 
\code{\link{acqfmhsp}},
\code{\link{acqhisp}},
\code{\link{acqnhhsp}},
\code{\link{acqvahsp}},
\code{\link{acqwkhsp}},
\code{\link{comhisp}},
\code{\link{engteach}},
\code{\link{fairhsps}},
\code{\link{famhsps}},
\code{\link{farehsps}},
\code{\link{feelhsps}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hspasn10}},
\code{\link{hspgovt}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influhsp}},
\code{\link{intlhsps}},
\code{\link{knwhisp}},
\code{\link{livehsps}},
\code{\link{marhisp}},
\code{\link{patrhsps}},
\code{\link{trthisp}},
\code{\link{ushisp}},
\code{\link{wlthhsps}},
\code{\link{workhsps}}
}
\concept{Hispanics}
\concept{Intergroup Relations}
\keyword{variable}
