% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_331.R
\name{violpeop}
\alias{violpeop}
\title{Would child do something violent toward other}
\source{
General Social Survey https://gss.norc.org
}
\description{
violpeop
}
\details{
Question A. How likely in your opinion is it (John/Mary) would do something violent toward other people -- is it very likely, somewhat likely, not very likely, or not at all likely?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2955/vshow}{GSS Data Explorer page for \code{violpeop}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not likely at all \tab don't know \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2002 \tab 1372 \tab 64 \tab 266 \tab 613 \tab 389 \tab 48 \tab 13 \tab - \tab - \tab 2765 \cr
   2024 \tab 1663 \tab 83 \tab 424 \tab 780 \tab 286 \tab 43 \tab 6 \tab - \tab 24 \tab 3309 \cr
   Total \tab 3035 \tab 147 \tab 690 \tab 1393 \tab 675 \tab 91 \tab 19 \tab 0 \tab 24 \tab 6074 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{comknows}},
\code{\link{failure}},
\code{\link{illphys}},
\code{\link{medkdneg}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{putsoff}},
\code{\link{sufadult}},
\code{\link{violself}}

Other Children: 
\code{\link{comknows}},
\code{\link{failure}},
\code{\link{illphys}},
\code{\link{medkdneg}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{sufadult}},
\code{\link{violself}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
