% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_331.R
\name{violself}
\alias{violself}
\title{Would child do something violent to own self}
\source{
General Social Survey https://gss.norc.org
}
\description{
violself
}
\details{
Question B. How likely in your opinion is it (John/Mary) would do something violent to (him/her)self?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2956/vshow}{GSS Data Explorer page for \code{violself}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 167 \tab 323 \tab 515 \tab 324 \tab 52 \tab 12 \tab 2765 \cr
   Total \tab 1372 \tab 167 \tab 323 \tab 515 \tab 324 \tab 52 \tab 12 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{comknows}},
\code{\link{failure}},
\code{\link{illphys}},
\code{\link{medkdneg}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{putsoff}},
\code{\link{sufadult}},
\code{\link{violpeop}}
}
\concept{Mental Health II}
\keyword{variable}
