% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_382.R
\name{visart}
\alias{visart}
\title{How often r visited art museum last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
visart
}
\details{
Question An art museum. How many times did you visit an art museum during the last year?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
   B/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3463/vshow}{GSS Data Explorer page for \code{visart}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 15 \tab 2 \tab 20 \tab 200 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab 75 \tab 8 \tab 9 \tab no answer \tab 24 \tab Total \cr
   2008 \tab 518 \tab 1032 \tab 220 \tab 9 \tab 1 \tab 8 \tab 1 \tab 106 \tab 3 \tab 1 \tab 56 \tab 1 \tab 26 \tab 12 \tab 15 \tab 7 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 2023 \cr
   2012 \tab 972 \tab 671 \tab 167 \tab 11 \tab - \tab 5 \tab 1 \tab 67 \tab 2 \tab - \tab 30 \tab 2 \tab 15 \tab 13 \tab 10 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 1974 \cr
   Total \tab 1490 \tab 1703 \tab 387 \tab 20 \tab 1 \tab 13 \tab 2 \tab 173 \tab 5 \tab 1 \tab 86 \tab 3 \tab 41 \tab 25 \tab 25 \tab 9 \tab 1 \tab 4 \tab 2 \tab 5 \tab 1 \tab 3997 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intenvir}},
\code{\link{intmil}},
\code{\link{intspace}},
\code{\link{sciintro}},
\code{\link{scimath}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
