% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{visions}
\alias{visions}
\title{Perceived events at great distance}
\source{
General Social Survey https://gss.norc.org
}
\description{
visions
}
\details{
Question How often have you had any of the following experiences?  READ EACH ITEM AND CIRCLE ONE CODE FOR EACH.
C. Seen events that happened at a great distance as they were happening.
}
\section{Values}{

\itemize{
\item \code{1} never in my life
\item \code{2} once or twice
\item \code{3} several times
\item \code{4} often
\item \code{8} cannot answer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/318/vshow}{GSS Data Explorer page for \code{visions}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never in my life \tab once or twice \tab several times \tab often \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 1009 \tab 254 \tab 132 \tab 39 \tab 20 \tab 19 \tab 1473 \cr
   1988 \tab - \tab 1033 \tab 248 \tab 110 \tab 49 \tab 27 \tab 14 \tab 1481 \cr
   1989 \tab 531 \tab 754 \tab 141 \tab 67 \tab 21 \tab 19 \tab 4 \tab 1537 \cr
   Total \tab 531 \tab 2796 \tab 643 \tab 309 \tab 109 \tab 66 \tab 37 \tab 4491 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}}
}
\concept{Core}
\keyword{variable}
