% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_149.R
\name{visitart}
\alias{visitart}
\title{Visited art museum or gallery in lst yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
visitart
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
B. Visit an art museum or gallery.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1406/vshow}{GSS Data Explorer page for \code{visitart}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 646 \tab 947 \tab 1 \tab 12 \tab 1606 \cr
   1998 \tab 1387 \tab 527 \tab 907 \tab 1 \tab 10 \tab 2832 \cr
   2002 \tab - \tab 1018 \tab 1721 \tab 14 \tab 12 \tab 2765 \cr
   Total \tab 1387 \tab 2191 \tab 3575 \tab 16 \tab 34 \tab 7203 \cr
}
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{conrock}},
\code{\link{hvymetal}},
\code{\link{makeart}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{rap}},
\code{\link{reggae}}
}
\concept{Culture}
\keyword{variable}
