% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{visitors}
\alias{visitors}
\title{Number of visitor's in household}
\source{
General Social Survey https://gss.norc.org
}
\description{
visitors
}
\details{
Question Number of visitors in household.
}
\section{Values}{

\itemize{
\item \verb{[0]} no visitors
\item \verb{[1]} 1 visitor
\item \verb{[2]} 2 visitors
\item \verb{[3]} 3 visitors
\item \verb{[4]} 4 visitors
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5438/vshow}{GSS Data Explorer page for \code{visitors}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no visitors \tab 1 visitor \tab 2 visitors \tab 4 visitors \tab 3 visitors \tab not available in this release \tab Total \cr
   1975 \tab - \tab 1454 \tab 25 \tab 8 \tab 3 \tab - \tab - \tab 1490 \cr
   1976 \tab 755 \tab 738 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1512 \tab 15 \tab 2 \tab 1 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1509 \tab 12 \tab 6 \tab 2 \tab 3 \tab - \tab 1532 \cr
   1980 \tab - \tab 1436 \tab 21 \tab 7 \tab 2 \tab 2 \tab - \tab 1468 \cr
   1982 \tab - \tab 1839 \tab 16 \tab 3 \tab - \tab 2 \tab - \tab 1860 \cr
   1983 \tab - \tab 1574 \tab 19 \tab 3 \tab - \tab 3 \tab - \tab 1599 \cr
   1984 \tab - \tab 1443 \tab 20 \tab 7 \tab - \tab 3 \tab - \tab 1473 \cr
   1985 \tab - \tab 1498 \tab 30 \tab 5 \tab 1 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1452 \tab 14 \tab 4 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1794 \tab 18 \tab 4 \tab 2 \tab 1 \tab - \tab 1819 \cr
   1988 \tab - \tab 1469 \tab 9 \tab 2 \tab - \tab 1 \tab - \tab 1481 \cr
   1989 \tab - \tab 1522 \tab 9 \tab 5 \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab - \tab 1359 \tab 10 \tab 3 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1498 \tab 14 \tab 5 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1597 \tab 8 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2970 \tab 17 \tab 4 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 2881 \tab 15 \tab 7 \tab - \tab 1 \tab - \tab 2904 \cr
   1998 \tab - \tab 2805 \tab 12 \tab 9 \tab 2 \tab 4 \tab - \tab 2832 \cr
   2000 \tab - \tab 2808 \tab 5 \tab 3 \tab - \tab 1 \tab - \tab 2817 \cr
   2002 \tab - \tab 2759 \tab 5 \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2798 \tab 14 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4462 \tab 36 \tab 10 \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab - \tab 2004 \tab 15 \tab 3 \tab - \tab 1 \tab - \tab 2023 \cr
   2010 \tab - \tab 2018 \tab 19 \tab 7 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1955 \tab 16 \tab 2 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2520 \tab 15 \tab 1 \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab - \tab 2839 \tab 23 \tab 3 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 2316 \tab 27 \tab 5 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 4299 \tab 58829 \tab 461 \tab 123 \tab 18 \tab 27 \tab 4032 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{famgen}},
\code{\link{hhtype}},
\code{\link{hhtype1}},
\code{\link{relhh1}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}}
}
\concept{Household Members and Structure}
\keyword{variable}
