% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{vistholy}
\alias{vistholy}
\title{How often r visits holy place for religious reasons}
\source{
General Social Survey https://gss.norc.org
}
\description{
vistholy
}
\details{
Question How often do you visit a holy place for religious reasons such as going to a shrine, temple, church, or mosque?
Please do not count attending regular religious services at your usual place of worship, if you have one.
Would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than once a year
\item \verb{[3]} about once or twice a year
\item \verb{[4]} several times a year
\item \verb{[5]} about once a month or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4710/vshow}{GSS Data Explorer page for \code{vistholy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than once a year \tab about once or twice a year \tab several times a year \tab about once a month or more \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 847 \tab 142 \tab 166 \tab 97 \tab 102 \tab 7 \tab 4 \tab 2023 \cr
   2018 \tab 1173 \tab 698 \tab 115 \tab 137 \tab 80 \tab 131 \tab 2 \tab 12 \tab 2348 \cr
   Total \tab 1831 \tab 1545 \tab 257 \tab 303 \tab 177 \tab 233 \tab 9 \tab 16 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{rightpeo}}
}
\concept{ISSP Religion}
\keyword{variable}
