% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{voedcol}
\alias{voedcol}
\title{Non-college postsecondary education (voednme1)}
\source{
General Social Survey https://gss.norc.org
}
\description{
voedcol
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
E. Besides attending college for credit, did you complete any courses for credit at an institution such as a business college, technical or vocational school?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/73/vshow}{GSS Data Explorer page for \code{voedcol}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab not available in this year \tab Total \cr
   2012 \tab 1008 \tab 287 \tab 678 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 1076 \tab 452 \tab 1009 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1215 \tab 522 \tab 1126 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 5647 \tab 1261 \tab 2813 \tab 6 \tab 0 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
