% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{voednme2}
\alias{voednme2}
\title{Postsecondary institution attended for credit}
\source{
General Social Survey https://gss.norc.org
}
\description{
voednme2
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
H. IF YES: What institution was that?
}
\section{Values}{

\itemize{
\item \code{1} jail/prison
\item \code{2} high school
\item \code{3} company/business
\item \code{4} vocational school/trades/technical (not high school, not college)
\item \code{5} secretarial school
\item \code{6} adult education/school
\item \code{7} medical/nursing/health/dental/hospital
\item \code{8} union
\item \code{9} armed forces
\item \code{10} college/university
\item \code{11} art/design
\item \code{12} business college (not business school at a university)
\item \code{13} home study/correspondence
\item \code{14} job corp
\item \code{15} beauty school/cosmetology
\item \code{16} insurance, real estate, finance, executive management
\item \code{17} police/fire academy; training
\item \code{18} development center
\item \code{19} career center
\item \code{20} foreign college/university
\item \code{30} other
\item \code{40} unable to code response, unclear, not identifiable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/76/vshow}{GSS Data Explorer page for \code{voednme2}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab high school \tab company/business \tab vocational school/trades/technical (not high school, not college) \tab adult education/school \tab medical/nursing/health/dental/hospital \tab armed forces \tab college/university \tab business college (not business school at a university) \tab home study/correspondence \tab job corp \tab beauty school/cosmetology \tab insurance, real estate, finance, executive management \tab police/fire academy; training \tab development center \tab career center \tab foreign college/university \tab other \tab uncodeable \tab secretarial school \tab union \tab jail/prison \tab art/design \tab unable to code response, unclear, not identifiable \tab Total \cr
   2012 \tab 1730 \tab 8 \tab 4 \tab 34 \tab 5 \tab 10 \tab 4 \tab 125 \tab 12 \tab 2 \tab 1 \tab 11 \tab 6 \tab 2 \tab 1 \tab 2 \tab 4 \tab 5 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2250 \tab 3 \tab 3 \tab 65 \tab 2 \tab 11 \tab 7 \tab 119 \tab 12 \tab 2 \tab 5 \tab 14 \tab 2 \tab 1 \tab 2 \tab 1 \tab 6 \tab 23 \tab 6 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2517 \tab 18 \tab 9 \tab 51 \tab - \tab 13 \tab 12 \tab 155 \tab 9 \tab 2 \tab 4 \tab 23 \tab 4 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 11 \tab 3 \tab 6 \tab 1 \tab 1 \tab 19 \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 8845 \tab 29 \tab 16 \tab 150 \tab 7 \tab 34 \tab 23 \tab 399 \tab 33 \tab 6 \tab 10 \tab 48 \tab 12 \tab 5 \tab 3 \tab 5 \tab 14 \tab 29 \tab 25 \tab 4 \tab 9 \tab 1 \tab 1 \tab 19 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}}
}
\concept{Core}
\keyword{variable}
