% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{volactyr}
\alias{volactyr}
\title{Since last yr any volunteering}
\source{
General Social Survey https://gss.norc.org
}
\description{
volactyr
}
\details{
Question Since (CURRENT MONTH) 1st of last year, have you done any volunteer activities through or for an organization?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3970/vshow}{GSS Data Explorer page for \code{volactyr}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{prespop}},
\code{\link{volacty2}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
