% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{volarts}
\alias{volarts}
\title{Volunteered for organizations of arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
volarts
}
\details{
Question Last year, did you do any volunteer work for organizations involved in the arts, culture, or the humanities?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2133/vshow}{GSS Data Explorer page for \code{volarts}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{drama}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{stateart}}

Other Leisure: 
\code{\link{drama}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{relart}}
}
\concept{Leisure}
\concept{Religion and Culture}
\concept{Voluntary Associations}
\keyword{variable}
