% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_200.R
\name{volhuman}
\alias{volhuman}
\title{Done volunteer work in human services}
\source{
General Social Survey https://gss.norc.org
}
\description{
volhuman
}
\details{
Question Listed on this card are examples of the many different areas in which people do volunteer activity. By volunteer activity I mean not just belonging to a service organization, but actually working in some way to help others for no monetary pay. In which, if any, of the areas listed on this card have you done some volunteer work in the past twelve months? (DON'T READ. RECORD BELOW UNDER VOLHLTH-VOLOTH. CIRCLE "YES" OR "NO" FOR EACH AREA)
D. Human services
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1888/vshow}{GSS Data Explorer page for \code{volhuman}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 137 \tab 1218 \tab 89 \tab 2904 \cr
   Total \tab 1460 \tab 137 \tab 1218 \tab 89 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{voleduc}},
\code{\link{volenvir}},
\code{\link{volhlth}},
\code{\link{volpub}},
\code{\link{volrec}},
\code{\link{volrelig}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
