% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{volmonth}
\alias{volmonth}
\title{Volunteer in last month}
\source{
General Social Survey https://gss.norc.org
}
\description{
volmonth
}
\details{
Question In the past month, did you do any volunteer activity through organizations, that is donate your time and energy not for pay?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/C/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2894/vshow}{GSS Data Explorer page for \code{volmonth}}.
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{cutahead}},
\code{\link{directns}},
\code{\link{empathy1}},
\code{\link{empathy2}},
\code{\link{empathy3}},
\code{\link{empathy4}},
\code{\link{empathy5}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{givblood}},
\code{\link{givchrty}},
\code{\link{givhmlss}},
\code{\link{givseat}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{retchnge}},
\code{\link{selffrst}},
\code{\link{selfless}},
\code{\link{valgiven}},
\code{\link{volchrty}}
}
\concept{Altruism}
\keyword{variable}
