% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_200.R
\name{volpub}
\alias{volpub}
\title{Done volunteer work in pub-soc benefit}
\source{
General Social Survey https://gss.norc.org
}
\description{
volpub
}
\details{
Question Listed on this card are examples of the many different areas in which people do volunteer activity. By volunteer activity I mean not just belonging to a service organization, but actually working in some way to help others for no monetary pay. In which, if any, of the areas listed on this card have you done some volunteer work in the past twelve months? (DON'T READ. RECORD BELOW UNDER VOLHLTH-VOLOTH. CIRCLE "YES" OR "NO" FOR EACH AREA)
F. Public/society benefit
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1890/vshow}{GSS Data Explorer page for \code{volpub}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{voleduc}},
\code{\link{volenvir}},
\code{\link{volhlth}},
\code{\link{volhuman}},
\code{\link{volrec}},
\code{\link{volrelig}}

Other Voluntary Associations: 
\code{\link{voleduc}},
\code{\link{volenvir}},
\code{\link{volhlth}},
\code{\link{volhuman}},
\code{\link{volrec}},
\code{\link{volrelig}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\concept{Voluntary Associations}
\keyword{variable}
