% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{volwkrel}
\alias{volwkrel}
\title{Volunteered religious past 12 month}
\source{
General Social Survey https://gss.norc.org
}
\description{
volwkrel
}
\details{
Question Have you done any voluntary activity in the past 12 months in any of the following areas? Voluntary activity is unpaid work, not just belonging to an organization or group. It should be of service or benefit to other people or the community and not only to one's family or personal friends. During the last 12 months did you do volunteer work in any of the following areas:
C. Religious and church related activities (helping churches and religious groups)
}
\section{Values}{

\itemize{
\item \code{1} no
\item \code{2} yes 1-2 times
\item \code{3} yes 3-5 times
\item \code{4} yes 6/more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4693/vshow}{GSS Data Explorer page for \code{volwkrel}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes 1-2 times \tab yes 3-5 times \tab yes 6/more times \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 787 \tab 191 \tab 85 \tab 163 \tab 58 \tab 2832 \cr
   Total \tab 1548 \tab 787 \tab 191 \tab 85 \tab 163 \tab 58 \tab 2832 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}}
}
\concept{ISSP Religion}
\keyword{variable}
