% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{vote08}
\alias{vote08}
\title{Did r vote in 2008 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote08
}
\details{
Question K. In 2008, you remember that Obama ran for President on the Democratic ticket against McCain for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} ineligible
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/172/vshow}{GSS Data Explorer page for \code{vote08}}.
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote04}}
}
\concept{Core}
\keyword{variable}
