% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{vote12}
\alias{vote12}
\title{Did r vote in 2012 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote12
}
\details{
Question L. In 2012, you remember that Obama ran for President on the Democratic ticket against Romney for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \verb{[1]} voted
\item \verb{[2]} did not vote
\item \verb{[3]} ineligible
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/175/vshow}{GSS Data Explorer page for \code{vote12}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab voted \tab did not vote \tab ineligible \tab don't know \tab no answer \tab Total \cr
   2014 \tab - \tab 1652 \tab 722 \tab 135 \tab 23 \tab 6 \tab 2538 \cr
   2016 \tab - \tab 1809 \tab 800 \tab 201 \tab 48 \tab 9 \tab 2867 \cr
   2018 \tab - \tab 1488 \tab 565 \tab 252 \tab 33 \tab 10 \tab 2348 \cr
   Total \tab 0 \tab 4949 \tab 2087 \tab 588 \tab 104 \tab 25 \tab 7753 \cr
}
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{pres12}}
}
\concept{Core}
\keyword{variable}
