% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{vote16}
\alias{vote16}
\title{Remember if voted in 2016 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote16
}
\details{
Question In 2016, you remember that Clinton ran for President on the Democratic ticket against Trump for the Republicans.
Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} ineligible
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7089/vshow}{GSS Data Explorer page for \code{vote16}}.
}

\keyword{variable}
