% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{vote20}
\alias{vote20}
\title{Remember if voted in 2020 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote20
}
\details{
Question In 2020, you remember that Joe Biden ran for President on the Democratic ticket against Donald Trump for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} ineligible
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7758/vshow}{GSS Data Explorer page for \code{vote20}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab ineligible \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 2474 \tab 681 \tab 166 \tab 87 \tab 130 \tab 6 \tab 3544 \cr
   Total \tab 0 \tab 2474 \tab 681 \tab 166 \tab 87 \tab 130 \tab 6 \tab 3544 \cr
}
}

\keyword{variable}
