% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{vote68}
\alias{vote68}
\title{Did r vote in 1968 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote68
}
\details{
Question A. Now in 1968, you remember that Humphrey ran for President on the Democratic ticket against Nixon for the Republicans, and Wallace as an Independent. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \verb{[1]} voted
\item \verb{[2]} did not vote
\item \verb{[3]} not eligible
\item \verb{[4]} refused
\item \verb{[6]} wldnt vt-relig
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/142/vshow}{GSS Data Explorer page for \code{vote68}}.

Counts by year:\tabular{llllllllll}{
   year \tab voted \tab did not vote \tab not eligible \tab refused \tab don't know \tab wldnt vt-relig \tab no answer \tab iap \tab Total \cr
   1972 \tab 1062 \tab 391 \tab 134 \tab 1 \tab 25 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 930 \tab 422 \tab 121 \tab 3 \tab 21 \tab 2 \tab 5 \tab - \tab 1504 \cr
   Total \tab 1992 \tab 813 \tab 255 \tab 4 \tab 46 \tab 2 \tab 5 \tab 0 \tab 3117 \cr
}
}

\seealso{
Other Core: 
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{partyid}},
\code{\link{pres68}},
\code{\link{spdotpre}}
}
\concept{Core}
\keyword{variable}
