% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{vote88}
\alias{vote88}
\title{Did r vote in 1988 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote88
}
\details{
Question F. In 1988, you remember that Dukakis ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \verb{[1]} voted
\item \verb{[2]} did not vote
\item \verb{[3]} not eligible
\item \verb{[4]} refused
\item \verb{[6]} wldnt vt-relig
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/157/vshow}{GSS Data Explorer page for \code{vote88}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab not eligible \tab refused \tab don't know \tab no answer \tab Total \cr
   1989 \tab - \tab 986 \tab 490 \tab 35 \tab 18 \tab 8 \tab - \tab 1537 \cr
   1990 \tab - \tab 888 \tab 410 \tab 48 \tab 8 \tab 17 \tab 1 \tab 1372 \cr
   1991 \tab - \tab 982 \tab 424 \tab 69 \tab 14 \tab 25 \tab 3 \tab 1517 \cr
   1993 \tab - \tab 1007 \tab 441 \tab 117 \tab 5 \tab 30 \tab 6 \tab 1606 \cr
   Total \tab 0 \tab 3863 \tab 1765 \tab 269 \tab 45 \tab 80 \tab 10 \tab 6032 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
