% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{vote92}
\alias{vote92}
\title{Did r vote in 1992 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote92
}
\details{
Question G. In 1992, you remember that Clinton ran for President on the Democratic ticket against Bush for the Republicans and Perot as an Independent.
Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} not eligible
\item \code{4} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/160/vshow}{GSS Data Explorer page for \code{vote92}}.
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
