% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{vote96}
\alias{vote96}
\title{Did r vote in 1996 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote96
}
\details{
Question H. In 1996, you remember that Clinton ran for President on the Democratic ticket against Dole for the Republicans and Perot for the Reform Party. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} ineligible
\item \code{4} refused to answer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/163/vshow}{GSS Data Explorer page for \code{vote96}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab ineligible \tab refused to answer \tab don't know \tab no answer \tab Total \cr
   1998 \tab - \tab 1783 \tab 830 \tab 123 \tab 15 \tab 51 \tab 30 \tab 2832 \cr
   2000 \tab - \tab 1737 \tab 799 \tab 201 \tab 9 \tab 64 \tab 7 \tab 2817 \cr
   2002 \tab - \tab 1734 \tab 657 \tab 340 \tab 13 \tab 21 \tab - \tab 2765 \cr
   Total \tab 0 \tab 5254 \tab 2286 \tab 664 \tab 37 \tab 136 \tab 37 \tab 8414 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}}
}
\concept{Core}
\keyword{variable}
