% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{voteprob}
\alias{voteprob}
\title{Voting problem in us elections}
\source{
General Social Survey https://gss.norc.org
}
\description{
voteprob
}
\details{
Question If you had to choose, which of the following do you think is a bigger problem for U.S. elections? Ineligible voters being allowed to vote, or eligible voters not being allowed to vote?
}
\section{Values}{

\itemize{
\item \verb{[1]} very strongly ineligible voters being allowed to vote
\item \verb{[2]} moderately strongly ineligible voters being allowed to vote
\item \verb{[3]} a little ineligible voters being allowed to vote
\item \verb{[4]} a little eligible voters not being allowed to vote
\item \verb{[5]} moderately strongly eligible voters not being allowed to vote
\item \verb{[6]} very strongly eligible voters not being allowed to vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8191/vshow}{GSS Data Explorer page for \code{voteprob}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab very strongly ineligible voters being allowed to vote \tab moderately strongly ineligible voters being allowed to vote \tab a little ineligible voters being allowed to vote \tab a little eligible voters not being allowed to vote \tab moderately strongly eligible voters not being allowed to vote \tab very strongly eligible voters not being allowed to vote \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 776 \tab 452 \tab 230 \tab 297 \tab 636 \tab 608 \tab 205 \tab 17 \tab 22 \tab 66 \tab 3309 \cr
   Total \tab 0 \tab 776 \tab 452 \tab 230 \tab 297 \tab 636 \tab 608 \tab 205 \tab 17 \tab 22 \tab 66 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{aneseligible}},
\code{\link{anespost_status}},
\code{\link{anespre_status}},
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{whovote24}},
\code{\link{whovote24a}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
