% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_297.R
\name{wait1}
\alias{wait1}
\title{How long r wait treatment moderate activity}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait1
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own before seeking medical treatment?
A. Not being able to do moderate activities, such as moving a table, pushing a vacuum cleaner, bowling or playing golf?
}
\section{Values}{

\itemize{
\item \code{0} zero days
\item \code{991} some days
\item \code{992} some weeks
\item \code{993} some months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2699/vshow}{GSS Data Explorer page for \code{wait1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab zero days \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab Total \cr
   2000 \tab 2044 \tab 16 \tab 370 \tab 251 \tab 73 \tab 23 \tab 19 \tab 11 \tab 10 \tab 2817 \cr
   Total \tab 2044 \tab 16 \tab 370 \tab 251 \tab 73 \tab 23 \tab 19 \tab 11 \tab 10 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}}
}
\concept{Health Status}
\keyword{variable}
