% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_297.R
\name{wait2}
\alias{wait2}
\title{How long r wait treatment climbing stairs}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait2
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own before seeking medical treatment?
B. Not being able to climb several flights of stairs?
}
\section{Values}{

\itemize{
\item \code{0} zero days
\item \code{991} some days
\item \code{992} some weeks
\item \code{993} some months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2700/vshow}{GSS Data Explorer page for \code{wait2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab zero days \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab Total \cr
   2000 \tab 2057 \tab 29 \tab 403 \tab 227 \tab 63 \tab 14 \tab 13 \tab 5 \tab 6 \tab 2817 \cr
   Total \tab 2057 \tab 29 \tab 403 \tab 227 \tab 63 \tab 14 \tab 13 \tab 5 \tab 6 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{wait1}},
\code{\link{wait3}},
\code{\link{wait4}}
}
\concept{Health Status}
\keyword{variable}
