% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_297.R
\name{wait3}
\alias{wait3}
\title{How long r wait treatment doing less work p}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait3
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own
before seeking medical treatment?
C. Accomplishing less than you would like with your work or other regular daily activities as a result of your
physical health?
}
\section{Values}{

\itemize{
\item \verb{[0]} zero days
\item \verb{[991]} some days
\item \verb{[992]} some weeks
\item \verb{[993]} some months
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2701/vshow}{GSS Data Explorer page for \code{wait3}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab zero days \tab 1 \tab 10 \tab 14 \tab 15 \tab 18 \tab 2 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab 9 \tab some days \tab some weeks \tab some months \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1757 \tab 13 \tab 366 \tab 4 \tab 1 \tab 1 \tab 1 \tab 244 \tab 51 \tab 1 \tab 13 \tab 11 \tab 5 \tab 8 \tab 1 \tab 40 \tab 54 \tab 16 \tab 4 \tab 226 \tab 2817 \cr
   Total \tab 1757 \tab 13 \tab 366 \tab 4 \tab 1 \tab 1 \tab 1 \tab 244 \tab 51 \tab 1 \tab 13 \tab 11 \tab 5 \tab 8 \tab 1 \tab 40 \tab 54 \tab 16 \tab 4 \tab 226 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{wait1}},
\code{\link{wait2}},
\code{\link{wait4}}
}
\concept{Health Status}
\keyword{variable}
