% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_298.R
\name{wait8}
\alias{wait8}
\title{How long r wait treatment not calm and peace}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait8
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own before seeking medical treatment?
H. Not feeling calm and peaceful?
}
\section{Values}{

\itemize{
\item \code{0} zero days
\item \code{991} some days
\item \code{992} some weeks
\item \code{993} some months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2706/vshow}{GSS Data Explorer page for \code{wait8}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab zero days \tab 1 \tab 10 \tab 12 \tab 144 \tab 2 \tab 21 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab some days \tab some weeks \tab some months \tab no answer \tab Total \cr
   2000 \tab 2116 \tab 9 \tab 159 \tab 2 \tab 3 \tab 1 \tab 121 \tab 1 \tab 15 \tab 13 \tab 3 \tab 6 \tab 4 \tab 21 \tab 40 \tab 23 \tab 280 \tab 2817 \cr
   Total \tab 2116 \tab 9 \tab 159 \tab 2 \tab 3 \tab 1 \tab 121 \tab 1 \tab 15 \tab 13 \tab 3 \tab 6 \tab 4 \tab 21 \tab 40 \tab 23 \tab 280 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downqol}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait9}}
}
\concept{Health Status}
\keyword{variable}
