% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_298.R
\name{wait9}
\alias{wait9}
\title{How long r wait treatment lack of energy}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait9
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own before seeking medical treatment?
I. Not having lots of energy?
}
\section{Values}{

\itemize{
\item \verb{[0]} zero days
\item \verb{[991]} some days
\item \verb{[992]} some weeks
\item \verb{[993]} some months
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2707/vshow}{GSS Data Explorer page for \code{wait9}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab zero days \tab 1 \tab 10 \tab 12 \tab 14 \tab 2 \tab 24 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab some days \tab some weeks \tab some months \tab no answer \tab Total \cr
   2000 \tab 1987 \tab 15 \tab 231 \tab 2 \tab 1 \tab 1 \tab 157 \tab 1 \tab 35 \tab 2 \tab 15 \tab 5 \tab 8 \tab 4 \tab 22 \tab 44 \tab 26 \tab 261 \tab 2817 \cr
   Total \tab 1987 \tab 15 \tab 231 \tab 2 \tab 1 \tab 1 \tab 157 \tab 1 \tab 35 \tab 2 \tab 15 \tab 5 \tab 8 \tab 4 \tab 22 \tab 44 \tab 26 \tab 261 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downqol}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}}
}
\concept{Health Status}
\keyword{variable}
