% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{wantbest}
\alias{wantbest}
\title{R is sure that other ppl want the best for r}
\source{
General Social Survey https://gss.norc.org
}
\description{
wantbest
}
\details{
Question To what extent do you agree or disagree with the following statements?
B. Most of the time you can be sure other people want the best for you
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4237/vshow}{GSS Data Explorer page for \code{wantbest}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 158 \tab 542 \tab 268 \tab 131 \tab 25 \tab 14 \tab 11 \tab 2765 \cr
   Total \tab 1616 \tab 158 \tab 542 \tab 268 \tab 131 \tab 25 \tab 14 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{demands}},
\code{\link{helpfrds}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{usefrds}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
