% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{wantjob}
\alias{wantjob}
\title{Would r like a job now}
\source{
General Social Survey https://gss.norc.org
}
\description{
wantjob
}
\details{
Question Would you like to have a paid job now?
}
\section{Values}{

\itemize{
\item \code{1} want full
\item \code{2} want part
\item \code{3} want lt 10 hrs
\item \code{4} dont want
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4529/vshow}{GSS Data Explorer page for \code{wantjob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
