% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{wantjob1}
\alias{wantjob1}
\title{Which job would you prefer}
\source{
General Social Survey https://gss.norc.org
}
\description{
wantjob1
}
\details{
Question Suppose you could decide on your work situation at present. Which of the following would you prefer?
}
\section{Values}{

\itemize{
\item \verb{[1]} a full time job
\item \verb{[2]} a part-time job
\item \verb{[3]} a job with less than 10 hours a week
\item \verb{[4]} no paid job at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4577/vshow}{GSS Data Explorer page for \code{wantjob1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a full time job \tab a part-time job \tab a job with less than 10 hours a week \tab no paid job at all \tab no answer \tab don't know \tab Total \cr
   1998 \tab 1609 \tab 672 \tab 325 \tab 69 \tab 120 \tab 37 \tab - \tab 2832 \cr
   2006 \tab 2992 \tab 963 \tab 346 \tab 53 \tab 139 \tab 4 \tab 13 \tab 4510 \cr
   2016 \tab 1390 \tab 921 \tab 318 \tab 67 \tab 152 \tab 10 \tab 9 \tab 2867 \cr
   Total \tab 5991 \tab 2556 \tab 989 \tab 189 \tab 411 \tab 51 \tab 22 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
