% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{waterfam}
\alias{waterfam}
\title{Water pollution danger to my family}
\source{
General Social Survey https://gss.norc.org
}
\description{
waterfam
}
\details{
Question B. And do you think that pollution of America's rivers, lakes, and streams is…
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4763/vshow}{GSS Data Explorer page for \code{waterfam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab Total \cr
   1993 \tab 49 \tab 374 \tab 445 \tab 541 \tab 113 \tab 15 \tab 33 \tab 36 \tab 1606 \cr
   1994 \tab 1606 \tab 294 \tab 408 \tab 488 \tab 117 \tab 18 \tab 28 \tab 33 \tab 2992 \cr
   Total \tab 1655 \tab 668 \tab 853 \tab 1029 \tab 230 \tab 33 \tab 61 \tab 69 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\keyword{variable}
