% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{watergen}
\alias{watergen}
\title{Water pollution danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
watergen
}
\details{
Question A. In general, do you think that pollution of America's rivers, lakes, and streams is...
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4762/vshow}{GSS Data Explorer page for \code{watergen}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 537 \tab 484 \tab 413 \tab 54 \tab 9 \tab 32 \tab 28 \tab 1606 \cr
   1994 \tab 1606 \tab 422 \tab 421 \tab 400 \tab 63 \tab 6 \tab 29 \tab 45 \tab 2992 \cr
   2000 \tab 1541 \tab 423 \tab 412 \tab 288 \tab 44 \tab 14 \tab 44 \tab 51 \tab 2817 \cr
   2010 \tab 614 \tab 426 \tab 549 \tab 353 \tab 51 \tab 10 \tab 35 \tab 6 \tab 2044 \cr
   2016 \tab 1956 \tab 382 \tab 328 \tab 159 \tab 22 \tab 6 \tab 12 \tab 2 \tab 2867 \cr
   2018 \tab 1563 \tab 346 \tab 262 \tab 141 \tab 23 \tab 5 \tab 5 \tab 3 \tab 2348 \cr
   Total \tab 7329 \tab 2536 \tab 2456 \tab 1754 \tab 257 \tab 50 \tab 157 \tab 135 \tab 14674 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{chemgen}},
\code{\link{indusgen}},
\code{\link{tempgen}}

Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{waterfam}}
}
\concept{ISSP Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
