% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{watergen1}
\alias{watergen1}
\title{Water pollution danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
watergen1
}
\details{
Question And do you think that pollution of America's rivers, lakes and streams is…
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7492/vshow}{GSS Data Explorer page for \code{watergen1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 591 \tab 691 \tab 442 \tab 51 \tab 5 \tab 52 \tab 185 \tab 12 \tab 4032 \cr
   Total \tab 2003 \tab 591 \tab 691 \tab 442 \tab 51 \tab 5 \tab 52 \tab 185 \tab 12 \tab 4032 \cr
}
}

\keyword{variable}
