% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_648.R
\name{webdocexp}
\alias{webdocexp}
\title{Information online helped me understand a doctor}
\source{
General Social Survey https://gss.norc.org
}
\description{
webdocexp
}
\details{
Question (To what extent do you agree or disagree with the following statements? During the past 12 months, information on the internet...) ...helped me understand what a doctor tried to explain to me.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7814/vshow}{GSS Data Explorer page for \code{webdocexp}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2406 \tab 70 \tab 503 \tab 288 \tab 153 \tab 46 \tab 70 \tab 5 \tab 3 \tab 3544 \cr
   Total \tab 2406 \tab 70 \tab 503 \tab 288 \tab 153 \tab 46 \tab 70 \tab 5 \tab 3 \tab 3544 \cr
}
}

\keyword{variable}
