% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_262.R
\name{webmob}
\alias{webmob}
\title{R uses home internet through mobile device}
\source{
General Social Survey https://gss.norc.org
}
\description{
webmob
}
\details{
Question B. Do you have access to the Internet or World Wide Web in your home through a Internet-enabled mobile device like a smart phone, PDA, or BlackBerry?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2380/vshow}{GSS Data Explorer page for \code{webmob}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 1755 \tab 55 \tab 232 \tab 1 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 1718 \tab 59 \tab 193 \tab 1 \tab 3 \tab - \tab 1974 \cr
   2014 \tab 2247 \tab 92 \tab 197 \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2490 \tab 148 \tab 225 \tab 1 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 2036 \tab 176 \tab 135 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 3778 \tab 179 \tab 66 \tab - \tab - \tab 9 \tab 4032 \cr
   2022 \tab 3114 \tab 327 \tab 103 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 17138 \tab 1036 \tab 1151 \tab 4 \tab 9 \tab 9 \tab 19347 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compuse}},
\code{\link{webtv}}
}
\concept{Information Society}
\keyword{variable}
