% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_262.R
\name{webtv}
\alias{webtv}
\title{R has internet via webtv}
\source{
General Social Survey https://gss.norc.org
}
\description{
webtv
}
\details{
Question A. Do you have access to the Internet or World Wide Web in your home through Web TV?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2379/vshow}{GSS Data Explorer page for \code{webtv}}.
}

\seealso{
Other Information Society: 
\code{\link{compuse}},
\code{\link{webmob}}
}
\concept{Information Society}
\keyword{variable}
