% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{webyr}
\alias{webyr}
\title{When r first began using web}
\source{
General Social Survey https://gss.norc.org
}
\description{
webyr
}
\details{
Question B. When did you first begin using the World Wide Web (for purposes other than email)?
}
\section{Values}{

\itemize{
\item \verb{[1]} this year
\item \verb{[2]} in the last year
\item \verb{[3]} 1-2 years ago
\item \verb{[4]} 2-3 years ago
\item \verb{[5]} 3-4 years ago
\item \verb{[6]} 4-5 years ago
\item \verb{[7]} 5 or more years ago
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2620/vshow}{GSS Data Explorer page for \code{webyr}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab this year \tab in the last year \tab 1-2 years ago \tab 2-3 years ago \tab 3-4 years ago \tab 4-5 years ago \tab 5 or more years ago \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1159 \tab 40 \tab 128 \tab 201 \tab 251 \tab 189 \tab 220 \tab 494 \tab 2 \tab 81 \tab 2765 \cr
   2004 \tab 992 \tab 26 \tab 44 \tab 97 \tab 122 \tab 176 \tab 235 \tab 997 \tab 4 \tab 119 \tab 2812 \cr
   Total \tab 2151 \tab 66 \tab 172 \tab 298 \tab 373 \tab 365 \tab 455 \tab 1491 \tab 6 \tab 200 \tab 5577 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emailyr}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webable}}
}
\concept{Information Society}
\keyword{variable}
