% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{weekly}
\alias{weekly}
\title{What is your weekly wage before tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
weekly
}
\details{
Question IF HOURLY WAGE:
C. What was your weekly wage, before any taxes or deductions?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2163/vshow}{GSS Data Explorer page for \code{weekly}}.
}

\seealso{
Other Jobs: 
\code{\link{daily}},
\code{\link{hourly}},
\code{\link{monthly}},
\code{\link{trusting}},
\code{\link{yearly}}
}
\concept{Jobs}
\keyword{variable}
