% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{weighing}
\alias{weighing}
\title{Which is the best method to report weight of leaf}
\source{
General Social Survey https://gss.norc.org
}
\description{
weighing
}
\details{
Question As part of a laboratory experiment, five students measured the weight of the same leaf four times. They recorded 20 slightly different weights. All of the work was done carefully and correctly. Their goal was to be as accurate as possible and reduce error in the experiment to a minimum. Which of the following is the best method to report the weight of the leaf?
}
\section{Values}{

\itemize{
\item \code{1} ask the teacher to weigh the leaf
\item \code{2} report the first measurement
\item \code{3} average all of the weights that were recorded
\item \code{4} average the highest and lowest weights recorded
\item \code{5} discard the lowest five weights
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3483/vshow}{GSS Data Explorer page for \code{weighing}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
