% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{weight}
\alias{weight}
\title{R weighs how much}
\source{
General Social Survey https://gss.norc.org
}
\description{
weight
}
\details{
Question About how much do you weigh without shoes? Reported in pounds.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2846/vshow}{GSS Data Explorer page for \code{weight}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 102 \tab 103 \tab 105 \tab 106 \tab 108 \tab 110 \tab 112 \tab 113 \tab 115 \tab 117 \tab 118 \tab 119 \tab 120 \tab 122 \tab 123 \tab 124 \tab 125 \tab 126 \tab 127 \tab 128 \tab 129 \tab 130 \tab 131 \tab 132 \tab 133 \tab 134 \tab 135 \tab 136 \tab 137 \tab 138 \tab 139 \tab 140 \tab 141 \tab 142 \tab 143 \tab 144 \tab 145 \tab 146 \tab 147 \tab 148 \tab 149 \tab 150 \tab 152 \tab 153 \tab 154 \tab 155 \tab 156 \tab 157 \tab 158 \tab 159 \tab 160 \tab 161 \tab 162 \tab 163 \tab 164 \tab 165 \tab 166 \tab 167 \tab 168 \tab 169 \tab 170 \tab 171 \tab 172 \tab 173 \tab 174 \tab 175 \tab 176 \tab 177 \tab 178 \tab 179 \tab 180 \tab 181 \tab 182 \tab 183 \tab 184 \tab 185 \tab 186 \tab 187 \tab 188 \tab 189 \tab 190 \tab 191 \tab 192 \tab 193 \tab 194 \tab 195 \tab 196 \tab 197 \tab 198 \tab 200 \tab 201 \tab 202 \tab 203 \tab 204 \tab 205 \tab 206 \tab 207 \tab 208 \tab 209 \tab 210 \tab 212 \tab 213 \tab 215 \tab 216 \tab 217 \tab 218 \tab 219 \tab 220 \tab 224 \tab 225 \tab 226 \tab 227 \tab 228 \tab 229 \tab 230 \tab 231 \tab 234 \tab 235 \tab 237 \tab 239 \tab 240 \tab 242 \tab 244 \tab 245 \tab 246 \tab 247 \tab 250 \tab 251 \tab 253 \tab 254 \tab 255 \tab 256 \tab 258 \tab 260 \tab 261 \tab 264 \tab 265 \tab 266 \tab 267 \tab 268 \tab 270 \tab 274 \tab 275 \tab 278 \tab 280 \tab 283 \tab 285 \tab 290 \tab 296 \tab 300 \tab 315 \tab 319 \tab 320 \tab 323 \tab 330 \tab 345 \tab 350 \tab 356 \tab 365 \tab 410 \tab 90 \tab 92 \tab 97 \tab 98 \tab don't know \tab no answer \tab 100 \tab 107 \tab 111 \tab 114 \tab 116 \tab 121 \tab 151 \tab 214 \tab 221 \tab 222 \tab 223 \tab 232 \tab 233 \tab 238 \tab 269 \tab 289 \tab 292 \tab 295 \tab 308 \tab 310 \tab 311 \tab 332 \tab 338 \tab 340 \tab 342 \tab 344 \tab 380 \tab 395 \tab 85 \tab 101 \tab 109 \tab 199 \tab 248 \tab 257 \tab 262 \tab 271 \tab 273 \tab 276 \tab 281 \tab 297 \tab 299 \tab 305 \tab 306 \tab 325 \tab 326 \tab 327 \tab 335 \tab 359 \tab 360 \tab 369 \tab 375 \tab 400 \tab 425 \tab 94 \tab 96 \tab 99 \tab skipped on web \tab not available in this year \tab Total \cr
   2014 \tab 1288 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 8 \tab 2 \tab 4 \tab 6 \tab 1 \tab 1 \tab 1 \tab 13 \tab 3 \tab 2 \tab 3 \tab 14 \tab 3 \tab 2 \tab 4 \tab 3 \tab 21 \tab 3 \tab 3 \tab 1 \tab 3 \tab 34 \tab 4 \tab 3 \tab 7 \tab 3 \tab 35 \tab 3 \tab 6 \tab 4 \tab 1 \tab 24 \tab 1 \tab 11 \tab 4 \tab 3 \tab 59 \tab 4 \tab 5 \tab 3 \tab 26 \tab 4 \tab 4 \tab 5 \tab 1 \tab 51 \tab 3 \tab 8 \tab 2 \tab 6 \tab 46 \tab 1 \tab 2 \tab 13 \tab 2 \tab 51 \tab 1 \tab 3 \tab 2 \tab 4 \tab 33 \tab 2 \tab 4 \tab 6 \tab 1 \tab 68 \tab 1 \tab 5 \tab 5 \tab 1 \tab 27 \tab 8 \tab 4 \tab 3 \tab 1 \tab 40 \tab 1 \tab 2 \tab 3 \tab 1 \tab 30 \tab 2 \tab 2 \tab 2 \tab 50 \tab 2 \tab 5 \tab 1 \tab 3 \tab 20 \tab 2 \tab 2 \tab 4 \tab 1 \tab 25 \tab 7 \tab 3 \tab 20 \tab 3 \tab 3 \tab 6 \tab 1 \tab 31 \tab 1 \tab 13 \tab 3 \tab 1 \tab 2 \tab 1 \tab 20 \tab 2 \tab 1 \tab 7 \tab 1 \tab 1 \tab 18 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 16 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 11 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 5 \tab 1 \tab 7 \tab 1 \tab 5 \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 20 \tab 53 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab - \tab 1 \tab 6 \tab 1 \tab - \tab 5 \tab 1 \tab 2 \tab 11 \tab 3 \tab 4 \tab 1 \tab 21 \tab - \tab 3 \tab 3 \tab 23 \tab 3 \tab 3 \tab 3 \tab 3 \tab 42 \tab - \tab 3 \tab 2 \tab 5 \tab 42 \tab 2 \tab 2 \tab 9 \tab 4 \tab 52 \tab 1 \tab 8 \tab 4 \tab 3 \tab 33 \tab - \tab 5 \tab 5 \tab 2 \tab 53 \tab 5 \tab 6 \tab 7 \tab 28 \tab 6 \tab 3 \tab 5 \tab 2 \tab 50 \tab 4 \tab 3 \tab 5 \tab 4 \tab 41 \tab - \tab 4 \tab 5 \tab 3 \tab 55 \tab 2 \tab 10 \tab 3 \tab 3 \tab 35 \tab 1 \tab 2 \tab 11 \tab 1 \tab 61 \tab 2 \tab 3 \tab 6 \tab 2 \tab 49 \tab 5 \tab 3 \tab 6 \tab 1 \tab 53 \tab - \tab 5 \tab 3 \tab 2 \tab 37 \tab 4 \tab 2 \tab 6 \tab 59 \tab 5 \tab 2 \tab 5 \tab - \tab 18 \tab 3 \tab 4 \tab 5 \tab 3 \tab 26 \tab 2 \tab 3 \tab 29 \tab 2 \tab 2 \tab 3 \tab 2 \tab 36 \tab 3 \tab 24 \tab - \tab - \tab 2 \tab 1 \tab 21 \tab 1 \tab - \tab 11 \tab 2 \tab - \tab 18 \tab - \tab 2 \tab 6 \tab 2 \tab 1 \tab 25 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 18 \tab - \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 10 \tab - \tab 3 \tab - \tab 12 \tab 1 \tab 1 \tab 8 \tab - \tab 8 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 11 \tab 57 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 1557 \tab 1 \tab 4 \tab 6 \tab - \tab 2 \tab 14 \tab 4 \tab - \tab 15 \tab 5 \tab 5 \tab - \tab 23 \tab 2 \tab 3 \tab 6 \tab 30 \tab 6 \tab 6 \tab 6 \tab 3 \tab 53 \tab 1 \tab 6 \tab 1 \tab 5 \tab 41 \tab 2 \tab 7 \tab 5 \tab 2 \tab 43 \tab 2 \tab 5 \tab 5 \tab 2 \tab 37 \tab 7 \tab 8 \tab 9 \tab 2 \tab 60 \tab 11 \tab 5 \tab 7 \tab 46 \tab 4 \tab 6 \tab 6 \tab 4 \tab 78 \tab 2 \tab 3 \tab 6 \tab 3 \tab 57 \tab 8 \tab 10 \tab 7 \tab 8 \tab 80 \tab 3 \tab 11 \tab 3 \tab 5 \tab 60 \tab 5 \tab 5 \tab 10 \tab 5 \tab 93 \tab 1 \tab 6 \tab 5 \tab 4 \tab 47 \tab 3 \tab 10 \tab 2 \tab 4 \tab 71 \tab 3 \tab 6 \tab 4 \tab 6 \tab 39 \tab 6 \tab 3 \tab 14 \tab 87 \tab 4 \tab 4 \tab 5 \tab 4 \tab 20 \tab 4 \tab 2 \tab 4 \tab 2 \tab 40 \tab 4 \tab 2 \tab 33 \tab 4 \tab 3 \tab 4 \tab 1 \tab 58 \tab 3 \tab 23 \tab 1 \tab - \tab 4 \tab 5 \tab 44 \tab - \tab 3 \tab 10 \tab 2 \tab 1 \tab 25 \tab 3 \tab 1 \tab 10 \tab - \tab 1 \tab 42 \tab - \tab - \tab 1 \tab 8 \tab 2 \tab - \tab 16 \tab - \tab 2 \tab 13 \tab - \tab - \tab 2 \tab 14 \tab - \tab 7 \tab - \tab 17 \tab 2 \tab 3 \tab 7 \tab 1 \tab 12 \tab 3 \tab - \tab 6 \tab - \tab 6 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 16 \tab 43 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 38 \tab - \tab 3544 \cr
   Total \tab 3745 \tab 2 \tab 8 \tab 15 \tab 2 \tab 4 \tab 27 \tab 7 \tab 6 \tab 32 \tab 9 \tab 10 \tab 2 \tab 57 \tab 5 \tab 8 \tab 12 \tab 67 \tab 12 \tab 11 \tab 13 \tab 9 \tab 116 \tab 4 \tab 12 \tab 4 \tab 13 \tab 117 \tab 8 \tab 12 \tab 21 \tab 9 \tab 130 \tab 6 \tab 19 \tab 13 \tab 6 \tab 94 \tab 8 \tab 24 \tab 18 \tab 7 \tab 172 \tab 20 \tab 16 \tab 17 \tab 100 \tab 14 \tab 13 \tab 16 \tab 7 \tab 179 \tab 9 \tab 14 \tab 13 \tab 13 \tab 144 \tab 9 \tab 16 \tab 25 \tab 13 \tab 186 \tab 6 \tab 24 \tab 8 \tab 12 \tab 128 \tab 8 \tab 11 \tab 27 \tab 7 \tab 222 \tab 4 \tab 14 \tab 16 \tab 7 \tab 123 \tab 16 \tab 17 \tab 11 \tab 6 \tab 164 \tab 4 \tab 13 \tab 10 \tab 9 \tab 106 \tab 12 \tab 7 \tab 22 \tab 196 \tab 11 \tab 11 \tab 11 \tab 7 \tab 58 \tab 9 \tab 8 \tab 13 \tab 6 \tab 91 \tab 13 \tab 8 \tab 82 \tab 9 \tab 8 \tab 13 \tab 4 \tab 125 \tab 7 \tab 60 \tab 4 \tab 1 \tab 8 \tab 7 \tab 85 \tab 3 \tab 4 \tab 28 \tab 5 \tab 2 \tab 61 \tab 4 \tab 4 \tab 21 \tab 3 \tab 4 \tab 83 \tab 1 \tab 2 \tab 2 \tab 12 \tab 5 \tab 1 \tab 45 \tab 1 \tab 3 \tab 24 \tab 2 \tab 1 \tab 4 \tab 28 \tab 1 \tab 15 \tab 1 \tab 36 \tab 4 \tab 9 \tab 18 \tab 2 \tab 24 \tab 5 \tab 1 \tab 8 \tab 1 \tab 11 \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 47 \tab 153 \tab 4 \tab 5 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab 4 \tab 2 \tab 5 \tab 4 \tab 4 \tab 2 \tab 2 \tab 2 \tab 4 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 38 \tab 0 \tab 8430 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{chngtme}},
\code{\link{condemnd}},
\code{\link{cowrkhlp}},
\code{\link{cowrkint}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{fairearn}},
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{fringeok}},
\code{\link{handmove}},
\code{\link{haveinfo}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{hlpequip}},
\code{\link{hlthdays}},
\code{\link{hrsrelax}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{hyperten}},
\code{\link{jobfind1}},
\code{\link{jobsecok}},
\code{\link{knowschd}},
\code{\link{knowschd1}},
\code{\link{knowwhat}},
\code{\link{laidoff}},
\code{\link{learnnew}},
\code{\link{lotofsay}},
\code{\link{manvsemp}},
\code{\link{misswork}},
\code{\link{mntlhlth}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{myskills}},
\code{\link{ntwkhard}},
\code{\link{opdevel}},
\code{\link{overwork}},
\code{\link{painarms}},
\code{\link{partteam}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{prodctiv}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{respect}},
\code{\link{rincblls}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{satjob1}},
\code{\link{secondwk}},
\code{\link{setthngs}},
\code{\link{slpprblm}},
\code{\link{spvtrfair}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{suphelp}},
\code{\link{teamsafe}},
\code{\link{toofewwk}},
\code{\link{trainops}},
\code{\link{trdunion}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{usedup}},
\code{\link{usetech}},
\code{\link{waypaid}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbonus}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkpraise}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksmooth}},
\code{\link{wktopsat}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrktime}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Health: 
\code{\link{aidscndm}},
\code{\link{aidslook}},
\code{\link{anscitst}},
\code{\link{anydoc}},
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{crelesse}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{diffcare}},
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld4}},
\code{\link{disabld5}},
\code{\link{disabld6}},
\code{\link{disabld7}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc15a}},
\code{\link{doc16}},
\code{\link{doc16a}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc18a}},
\code{\link{doc19}},
\code{\link{doc19a}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc20a}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{docaskme}},
\code{\link{doccosts}},
\code{\link{docdecid}},
\code{\link{docinfo}},
\code{\link{doclist}},
\code{\link{docnomri}},
\code{\link{docrely}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{docvig}},
\code{\link{downblue}},
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downfeel}},
\code{\link{downqol}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotfeel}},
\code{\link{emotqol}},
\code{\link{emotrely}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{energy}},
\code{\link{findnwjb}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{health}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{helpsick}},
\code{\link{hivkiss}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest1_0622}},
\code{\link{hivtest2}},
\code{\link{hivtest2_0622}},
\code{\link{hivvac}},
\code{\link{hlth1}},
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{hlthall}},
\code{\link{hlthcare}},
\code{\link{hlthcovr}},
\code{\link{hlthdays}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthinsr}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthplan}},
\code{\link{hlthtv}},
\code{\link{hlthtype}},
\code{\link{hlthwww}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hrdshp6}},
\code{\link{hyperten}},
\code{\link{insrchng}},
\code{\link{insrlmts}},
\code{\link{intmed}},
\code{\link{limitedp}},
\code{\link{misswork}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{modact}},
\code{\link{mykids}},
\code{\link{nodocidu}},
\code{\link{othkids}},
\code{\link{othplan}},
\code{\link{pain}},
\code{\link{painarms}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}},
\code{\link{peaceful}},
\code{\link{physhlth}},
\code{\link{regdoc}},
\code{\link{regdocyr}},
\code{\link{sathealt}},
\code{\link{secopin}},
\code{\link{sexpaid}},
\code{\link{slpprblm}},
\code{\link{socacts}},
\code{\link{sppart}},
\code{\link{stairs}},
\code{\link{swithlth}},
\code{\link{treat1}},
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}},
\code{\link{treat6}},
\code{\link{treat7}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{usedmntl}},
\code{\link{usedup}},
\code{\link{wait1}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}}
}
\concept{Health}
\concept{Quality of Working Life}
\keyword{variable}
