% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{welfare1}
\alias{welfare1}
\title{Welfare makes people work less}
\source{
General Social Survey https://gss.norc.org
}
\description{
welfare1
}
\details{
Question Here are some opinions other people have expressed about welfare. For each of the following statements, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it.
Welfare . . . (READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.)
A. Makes people work less than they would if there wasn''t a welfare system.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5498/vshow}{GSS Data Explorer page for \code{welfare1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1986 \tab - \tab 481 \tab 753 \tab 191 \tab 27 \tab 17 \tab 1 \tab - \tab 1470 \cr
   Total \tab 0 \tab 481 \tab 753 \tab 191 \tab 27 \tab 17 \tab 1 \tab 0 \tab 1470 \cr
}
}

\seealso{
Other Welfare: 
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govunemp}},
\code{\link{lessfare}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}},
\code{\link{workfare}}

Other Income: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{earndes}},
\code{\link{earnmore}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{fairearn}},
\code{\link{famfinan}},
\code{\link{finalter}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{finrela}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{impfinan}},
\code{\link{incdef}},
\code{\link{inclstjb}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{justpay}},
\code{\link{minfood}},
\code{\link{minfour}},
\code{\link{mininc}},
\code{\link{minthree}},
\code{\link{othpay}},
\code{\link{owninc}},
\code{\link{pastpay}},
\code{\link{paychild}},
\code{\link{paychnge}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{payfam1}},
\code{\link{payhard}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{povline}},
\code{\link{realinc}},
\code{\link{realrinc}},
\code{\link{rincblls}},
\code{\link{rincom06}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{salsatfn}},
\code{\link{satfin}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}}

Other Welfare: 
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govunemp}},
\code{\link{lessfare}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}},
\code{\link{workfare}}
}
\concept{Income}
\concept{Welfare}
\keyword{variable}
