% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{welfare5}
\alias{welfare5}
\title{Helps prevent hunger and starvation}
\source{
General Social Survey https://gss.norc.org
}
\description{
welfare5
}
\details{
Question Here are some opinions other people have expressed about welfare. For each of the following statements, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it.
Welfare . . . (READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.)
E. Helps to prevent hunger and starvation.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5502/vshow}{GSS Data Explorer page for \code{welfare5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1986 \tab - \tab 228 \tab 1070 \tab 136 \tab 26 \tab 7 \tab 3 \tab 1470 \cr
   Total \tab 0 \tab 228 \tab 1070 \tab 136 \tab 26 \tab 7 \tab 3 \tab 1470 \cr
}
}

\seealso{
Other Welfare: 
\code{\link{welfare4}},
\code{\link{welfare6}}
}
\concept{Welfare}
\keyword{variable}
