% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{whatco2}
\alias{whatco2}
\title{Classify place where partners works}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatco2
}
\details{
Question How would you classify the place where your (husband/wife/spouse/partner) (worked/works) Was this mainly….?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} wholesale trade
\item \verb{[3]} retail trade
\item \verb{[4]} other (agriculture, construction, service, government, etc.)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8012/vshow}{GSS Data Explorer page for \code{whatco2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab don't know \tab iap \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab - \tab 52 \tab 11 \tab 62 \tab 217 \tab 2 \tab 3683 \tab 5 \tab - \tab 4032 \cr
   2022 \tab - \tab 40 \tab 6 \tab 44 \tab 192 \tab 7 \tab 3250 \tab 3 \tab 2 \tab 3544 \cr
   2024 \tab - \tab 33 \tab 10 \tab 33 \tab 163 \tab 13 \tab 3055 \tab 1 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 125 \tab 27 \tab 139 \tab 572 \tab 22 \tab 9988 \tab 9 \tab 3 \tab 10885 \cr
}
}

\keyword{variable}
