% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{whatma2}
\alias{whatma2}
\title{How would you classify the place where mother (or youngest same-sex parent) worked? Was this mainly...?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatma2
}
\details{
Question How would you classify the place where your mother worked? Was this mainly….?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} wholesale trade
\item \verb{[3]} retail trade
\item \verb{[4]} other (agriculture, construction, service, government, etc.)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7745/vshow}{GSS Data Explorer page for \code{whatma2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 276 \tab 44 \tab 402 \tab 1718 \tab 162 \tab 934 \tab 2 \tab 6 \tab 3544 \cr
   2024 \tab - \tab 251 \tab 32 \tab 338 \tab 1529 \tab 153 \tab 991 \tab 3 \tab 12 \tab 3309 \cr
   Total \tab 0 \tab 527 \tab 76 \tab 740 \tab 3247 \tab 315 \tab 1925 \tab 5 \tab 18 \tab 6853 \cr
}
}

\keyword{variable}
