% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{whatslf2}
\alias{whatslf2}
\title{How would you classify the place where you (#whatslf2fill)? Was this mainly...?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatslf2
}
\details{
Question How would you classify the place where you (work/worked)? Was this mainly...?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} wholesale trade
\item \verb{[3]} retail trade
\item \verb{[4]} other (agriculture, construction, service, government, etc.)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7746/vshow}{GSS Data Explorer page for \code{whatslf2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 228 \tab 62 \tab 283 \tab 1612 \tab 16 \tab 1326 \tab 12 \tab 5 \tab 3544 \cr
   2024 \tab - \tab 209 \tab 67 \tab 264 \tab 1429 \tab 10 \tab 1311 \tab 4 \tab 15 \tab 3309 \cr
   Total \tab 0 \tab 437 \tab 129 \tab 547 \tab 3041 \tab 26 \tab 2637 \tab 16 \tab 20 \tab 6853 \cr
}
}

\keyword{variable}
