% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{whatsp2}
\alias{whatsp2}
\title{How would you classify the place where your (#_spfill1) (#whatsp2fill)? Was this mainly...?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatsp2
}
\details{
Question How would you classify the place where your (spouse/partner) (work/worked)? Was this mainly...?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} wholesale trade
\item \verb{[3]} retail trade
\item \verb{[4]} other (agriculture, construction, service, government, etc.)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7748/vshow}{GSS Data Explorer page for \code{whatsp2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab don't know \tab iap \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab - \tab 247 \tab 56 \tab 244 \tab 1347 \tab 2 \tab 2117 \tab 19 \tab - \tab 4032 \cr
   2022 \tab - \tab 164 \tab 36 \tab 149 \tab 1022 \tab 31 \tab 2126 \tab 4 \tab 12 \tab 3544 \cr
   2024 \tab - \tab 157 \tab 31 \tab 158 \tab 910 \tab 31 \tab 2011 \tab 6 \tab 5 \tab 3309 \cr
   Total \tab 0 \tab 568 \tab 123 \tab 551 \tab 3279 \tab 64 \tab 6254 \tab 29 \tab 17 \tab 10885 \cr
}
}

\keyword{variable}
