% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_583.R
\name{where2}
\alias{where2}
\title{Where is 2nd person staying?}
\source{
General Social Survey https://gss.norc.org
}
\description{
where2
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
6. Where is the second person staying?
}
\section{Values}{

\itemize{
\item \code{1} another household
\item \code{2} traveling
\item \code{3} institution
\item \code{4} other, don't know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5343/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab another household \tab traveling \tab other, don't know \tab no answer \tab institution \tab not available in this year \tab don't know \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1479 \tab 11 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1495 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1522 \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1513 \tab 17 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1457 \tab 8 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1853 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1585 \tab 11 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1459 \tab 8 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1527 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1465 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1806 \tab 11 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1480 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1533 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 13 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1498 \tab 11 \tab 1 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1589 \tab 12 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2963 \tab 17 \tab 2 \tab - \tab 3 \tab 7 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2888 \tab 5 \tab 4 \tab - \tab 4 \tab 3 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2816 \tab 7 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2812 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2781 \tab 14 \tab 7 \tab 4 \tab - \tab 6 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4471 \tab 16 \tab 5 \tab 7 \tab - \tab 11 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1992 \tab 9 \tab 8 \tab 1 \tab 8 \tab 5 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2017 \tab 7 \tab 9 \tab 2 \tab 2 \tab 7 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1955 \tab 5 \tab 6 \tab 5 \tab - \tab 3 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2495 \tab 8 \tab 16 \tab 2 \tab - \tab 17 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2825 \tab 14 \tab 16 \tab 1 \tab 2 \tab 9 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2321 \tab 19 \tab 4 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 3494 \tab 19 \tab 22 \tab - \tab - \tab 6 \tab - \tab 3 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 67815 \tab 267 \tab 126 \tab 30 \tab 32 \tab 85 \tab 4032 \tab 3 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{where3}}

Other Family: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{where3}}

Other Household: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{where3}}
}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
