% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{wherewrk}
\alias{wherewrk}
\title{Doing your job in one or several different places}
\source{
General Social Survey https://gss.norc.org
}
\description{
wherewrk
}
\details{
Question When you are doing your job, do you usually do it in one place (like the same street address) or in several different places?
}
\section{Values}{

\itemize{
\item \code{1} some place
\item \code{2} different
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1372/vshow}{GSS Data Explorer page for \code{wherewrk}}.
}

\seealso{
Other Work Organizations: 
\code{\link{baseofop}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{splocnum}},
\code{\link{training}}

Other Job: 
\code{\link{baseofop}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{splocnum}},
\code{\link{training}}
}
\concept{Job}
\concept{Work Organizations}
\keyword{variable}
