% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{whichsci}
\alias{whichsci}
\title{Industry scientists less reliable than univer'sity scientists}
\source{
General Social Survey https://gss.norc.org
}
\description{
whichsci
}
\details{
Question I’m going to read to you some statements like those you might find in a newspaper or magazine article. For each statement, please tell me if you strongly agree, agree, disagree, or strongly disagree.
G. Research results from industry scientists are less reliable than those from university scientists.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3322/vshow}{GSS Data Explorer page for \code{whichsci}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}}

Other Science: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{seeksci}},
\code{\link{toofast}}
}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
