% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_605.R
\name{whoelse1}
\alias{whoelse1}
\title{Presence of others: children under six}
\source{
General Social Survey https://gss.norc.org
}
\description{
whoelse1
}
\details{
Question What other persons were present during the interview? Circle all that apply.
A. Children under 6
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5550/vshow}{GSS Data Explorer page for \code{whoelse1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1518 \tab 177 \tab 1254 \tab 8 \tab 35 \tab - \tab 2992 \cr
   1996 \tab - \tab 396 \tab 2430 \tab 7 \tab 71 \tab - \tab 2904 \cr
   1998 \tab - \tab 324 \tab 2387 \tab 4 \tab 117 \tab - \tab 2832 \cr
   2000 \tab - \tab 280 \tab 2389 \tab 2 \tab 146 \tab - \tab 2817 \cr
   2002 \tab - \tab 271 \tab 2479 \tab 15 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 280 \tab 2499 \tab 21 \tab 12 \tab - \tab 2812 \cr
   2006 \tab - \tab 413 \tab 4073 \tab 13 \tab 11 \tab - \tab 4510 \cr
   2008 \tab - \tab 224 \tab 1796 \tab - \tab 3 \tab - \tab 2023 \cr
   2010 \tab - \tab 231 \tab 1805 \tab - \tab 8 \tab - \tab 2044 \cr
   2012 \tab - \tab 237 \tab 1734 \tab - \tab 3 \tab - \tab 1974 \cr
   2014 \tab - \tab 249 \tab 2283 \tab - \tab 6 \tab - \tab 2538 \cr
   2016 \tab - \tab 273 \tab 2587 \tab - \tab 7 \tab - \tab 2867 \cr
   2018 \tab 9 \tab 241 \tab 2098 \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 1642 \tab 154 \tab 1654 \tab 94 \tab - \tab - \tab 3544 \cr
   Total \tab 3169 \tab 3750 \tab 31468 \tab 164 \tab 419 \tab 0 \tab 38970 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{whoelse2}}
}
\concept{Paradata}
\keyword{variable}
