% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_606.R
\name{whoelse3}
\alias{whoelse3}
\title{Presence of others: spouse partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
whoelse3
}
\details{
Question What other persons were present during the interview? Circle all that apply.
C. Spouse Partner
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5552/vshow}{GSS Data Explorer page for \code{whoelse3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1518 \tab 261 \tab 1170 \tab 8 \tab 35 \tab - \tab 2992 \cr
   1996 \tab - \tab 492 \tab 2334 \tab 7 \tab 71 \tab - \tab 2904 \cr
   1998 \tab - \tab 422 \tab 2289 \tab 4 \tab 117 \tab - \tab 2832 \cr
   2000 \tab - \tab 381 \tab 2288 \tab 2 \tab 146 \tab - \tab 2817 \cr
   2002 \tab - \tab 321 \tab 2429 \tab 15 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 343 \tab 2436 \tab 21 \tab 12 \tab - \tab 2812 \cr
   2006 \tab - \tab 543 \tab 3943 \tab 13 \tab 11 \tab - \tab 4510 \cr
   2008 \tab - \tab 286 \tab 1734 \tab - \tab 3 \tab - \tab 2023 \cr
   2010 \tab - \tab 326 \tab 1710 \tab - \tab 8 \tab - \tab 2044 \cr
   2012 \tab - \tab 322 \tab 1649 \tab - \tab 3 \tab - \tab 1974 \cr
   2014 \tab - \tab 435 \tab 2097 \tab - \tab 6 \tab - \tab 2538 \cr
   2016 \tab - \tab 471 \tab 2389 \tab - \tab 7 \tab - \tab 2867 \cr
   2018 \tab 9 \tab 339 \tab 2000 \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 1642 \tab 255 \tab 1553 \tab 94 \tab - \tab - \tab 3544 \cr
   Total \tab 3169 \tab 5197 \tab 30021 \tab 164 \tab 419 \tab 0 \tab 38970 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{intid}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}
}
\concept{Paradata}
\keyword{variable}
