% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_606.R
\name{whoelse4}
\alias{whoelse4}
\title{Presence of others: other relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
whoelse4
}
\details{
Question What other persons were present during the interview? Circle all that apply.
D. Other Relatives
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5553/vshow}{GSS Data Explorer page for \code{whoelse4}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1518 \tab 44 \tab 1387 \tab 8 \tab 35 \tab - \tab 2992 \cr
   1996 \tab - \tab 122 \tab 2704 \tab 7 \tab 71 \tab - \tab 2904 \cr
   1998 \tab - \tab 87 \tab 2624 \tab 4 \tab 117 \tab - \tab 2832 \cr
   2000 \tab - \tab 104 \tab 2565 \tab 2 \tab 146 \tab - \tab 2817 \cr
   2002 \tab - \tab 74 \tab 2676 \tab 15 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 77 \tab 2702 \tab 21 \tab 12 \tab - \tab 2812 \cr
   2006 \tab - \tab 135 \tab 4351 \tab 13 \tab 11 \tab - \tab 4510 \cr
   2008 \tab - \tab 92 \tab 1928 \tab - \tab 3 \tab - \tab 2023 \cr
   2010 \tab - \tab 98 \tab 1938 \tab - \tab 8 \tab - \tab 2044 \cr
   2012 \tab - \tab 89 \tab 1882 \tab - \tab 3 \tab - \tab 1974 \cr
   2014 \tab - \tab 126 \tab 2406 \tab - \tab 6 \tab - \tab 2538 \cr
   2016 \tab - \tab 119 \tab 2741 \tab - \tab 7 \tab - \tab 2867 \cr
   2018 \tab 9 \tab 121 \tab 2218 \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 1642 \tab 84 \tab 1724 \tab 94 \tab - \tab - \tab 3544 \cr
   Total \tab 3169 \tab 1372 \tab 33846 \tab 164 \tab 419 \tab 0 \tab 38970 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{intid}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse3}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}
}
\concept{Paradata}
\keyword{variable}
