% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{whosold}
\alias{whosold}
\title{Describe your purchase}
\source{
General Social Survey https://gss.norc.org
}
\description{
whosold
}
\details{
Question Which of the following best describes your purchase?  READ CATEGORIES BELOW.
}
\section{Values}{

\itemize{
\item \code{1} directly from builder
\item \code{2} directly from previous owner
\item \code{3} through a real estate agent
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1817/vshow}{GSS Data Explorer page for \code{whosold}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab directly from builder \tab directly from previous owner \tab through a real estate agent \tab other \tab Total \cr
   1996 \tab 2451 \tab 61 \tab 102 \tab 218 \tab 72 \tab 2904 \cr
   Total \tab 2451 \tab 61 \tab 102 \tab 218 \tab 72 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}}
}
\concept{Markets}
\keyword{variable}
