% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{whovote24}
\alias{whovote24}
\title{Voting intention in 2024 election (pre july 25 2024)}
\source{
General Social Survey https://gss.norc.org
}
\description{
whovote24
}
\details{
Question In the November election for President, do you think you will probably vote for Donald Trump, Joe Biden, someone else, or will you probably not vote?
}
\section{Values}{

\itemize{
\item \verb{[1]} donald trump
\item \verb{[2]} joe biden
\item \verb{[3]} someone else
\item \verb{[4]} probably not vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8193/vshow}{GSS Data Explorer page for \code{whovote24}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab donald trump \tab joe biden \tab someone else \tab probably not vote \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 538 \tab 559 \tab 275 \tab 270 \tab 50 \tab 1569 \tab 17 \tab 31 \tab 3309 \cr
   Total \tab 0 \tab 538 \tab 559 \tab 275 \tab 270 \tab 50 \tab 1569 \tab 17 \tab 31 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{aneseligible}},
\code{\link{anespre_status}},
\code{\link{ratechall1ts}},
\code{\link{ratechall2ts}},
\code{\link{ratechall3ts}},
\code{\link{rateincumbts}},
\code{\link{voteprob}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
