% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{whovote24a}
\alias{whovote24a}
\title{Voting intention in 2024 election (post july 25 2024)}
\source{
General Social Survey https://gss.norc.org
}
\description{
whovote24a
}
\details{
Question In the November election for President, do you think you will probably vote for Donald Trump, Kamala Harris, someone else, or will you probably not vote?
}
\section{Values}{

\itemize{
\item \verb{[1]} donald trump
\item \verb{[2]} kamala harris
\item \verb{[3]} someone else
\item \verb{[4]} probably not vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8192/vshow}{GSS Data Explorer page for \code{whovote24a}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab donald trump \tab kamala harris \tab someone else \tab probably not vote \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 541 \tab 576 \tab 90 \tab 247 \tab 40 \tab 1756 \tab 23 \tab 36 \tab 3309 \cr
   Total \tab 0 \tab 541 \tab 576 \tab 90 \tab 247 \tab 40 \tab 1756 \tab 23 \tab 36 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}}
}
\concept{ANES}
\keyword{variable}
