% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{whovotets}
\alias{whovotets}
\title{Voting intention - time series}
\source{
General Social Survey https://gss.norc.org
}
\description{
whovotets
}
\details{
Question (Recode of WHOVOTE24/WHOVOTE24A)
}
\section{Values}{

\itemize{
\item \verb{[1]} republican candidate
\item \verb{[2]} democratic candidate
\item \verb{[3]} someone else
\item \verb{[4]} probably not vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8194/vshow}{GSS Data Explorer page for \code{whovotets}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab republican candidate \tab democratic candidate \tab someone else \tab probably not vote \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 1079 \tab 1135 \tab 365 \tab 517 \tab 90 \tab 16 \tab 40 \tab 67 \tab 3309 \cr
   Total \tab 0 \tab 1079 \tab 1135 \tab 365 \tab 517 \tab 90 \tab 16 \tab 40 \tab 67 \tab 3309 \cr
}
}

\seealso{
Other ANES: 
\code{\link{aneseligible}},
\code{\link{anespost_status}},
\code{\link{anespre_status}},
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovote24a}}
}
\concept{ANES}
\keyword{variable}
