% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_256.R
\name{whtchng}
\alias{whtchng}
\title{Percentage change population of whites in 25 yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
whtchng
}
\details{
Question In the next 25 years, which of these groups do you think will increase their share of the population by a lot, which will increase by a little, which will stay the same, which will decrease by a little, and which will decrease by a lot?
A. Whites?
}
\section{Values}{

\itemize{
\item \code{1} increase a lot
\item \code{2} increase by a little
\item \code{3} stay the same
\item \code{4} decrease by a little
\item \code{5} decrease by a lot
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2317/vshow}{GSS Data Explorer page for \code{whtchng}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkchng}},
\code{\link{hispchng}},
\code{\link{jewschng}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
