% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_258.R
\name{whtschl}
\alias{whtschl}
\title{Does r know whites from school}
\source{
General Social Survey https://gss.norc.org
}
\description{
whtschl
}
\details{
Question Now, I'm going to ask some questions about these people you personally know.
A. Do you know any of these Whites from when you went to school or college?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2335/vshow}{GSS Data Explorer page for \code{whtschl}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2485 \tab 185 \tab 111 \tab 36 \tab 2817 \cr
   Total \tab 2485 \tab 185 \tab 111 \tab 36 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkschl}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtwrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
