% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{whybiz}
\alias{whybiz}
\title{Why did r start business}
\source{
General Social Survey https://gss.norc.org
}
\description{
whybiz
}
\details{
Question All in all, would you say you started, or are starting, your business because you saw an opportunity or you started it out of necessity?
IF R HAS MORE THAN ONE BUSINESS/FARM/PARTNERSHIP, ASK ABOUT THE PRIMARY BUSINESS THEY ARE INVOLVED IN.
}
\section{Values}{

\itemize{
\item \code{1} you started it because you came across an opportunity
\item \code{2} you started it because it was a necessity
\item \code{3} both (volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3822/vshow}{GSS Data Explorer page for \code{whybiz}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab you started it because you came across an opportunity \tab you started it because it was a necessity \tab both (volunteered) \tab don't know \tab Total \cr
   2008 \tab 1773 \tab 157 \tab 59 \tab 31 \tab 3 \tab 2023 \cr
   Total \tab 1773 \tab 157 \tab 59 \tab 31 \tab 3 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{owninc}},
\code{\link{paidemps}}
}
\concept{Self Employment}
\keyword{variable}
