% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_342.R
\name{whychng3}
\alias{whychng3}
\title{Religious change experience3}
\source{
General Social Survey https://gss.norc.org
}
\description{
whychng3
}
\details{
Question Open-ended questions for how people who had undergone spiritual or religious change described their experiences.
C. Third Mention
}
\section{Values}{

\itemize{
\item \code{96} no mention
\item \code{97} nothing special happened
\item \code{99} no asnwer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3073/vshow}{GSS Data Explorer page for \code{whychng3}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab 25 \tab 28 \tab 29 \tab 30 \tab 31 \tab 35 \tab 38 \tab 39 \tab 43 \tab 44 \tab 45 \tab 5 \tab 68 \tab 70 \tab 74 \tab 95 \tab no mention \tab nothing special happened \tab no answer \tab Total \cr
   2004 \tab 2163 \tab 1 \tab 3 \tab 3 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 584 \tab 1 \tab 32 \tab 2812 \cr
   Total \tab 2163 \tab 1 \tab 3 \tab 3 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 584 \tab 1 \tab 32 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{agerborn}},
\code{\link{chngrbrn}},
\code{\link{entity}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{numrborn}},
\code{\link{whychng1}},
\code{\link{whychng2}}
}
\concept{Spirituality and Religion}
\keyword{variable}
