% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{whyjbct1}
\alias{whyjbct1}
\title{Why job abolished or laid off}
\source{
General Social Survey https://gss.norc.org
}
\description{
whyjbct1
}
\details{
Question Why was the job abolished? Why were you permanently laid off?
}
\section{Values}{

\itemize{
\item \verb{[1]} company's sales were declining
\item \verb{[2]} machines and computers made my job obsolete
\item \verb{[3]} job requirements changed and i was not able to do the job
\item \verb{[4]} job moved to another location in the united states
\item \verb{[5]} job moved to another country
\item \verb{[6]} other
\item \verb{[7]} company sold
\item \verb{[8]} job done
\item \verb{[9]} company in financial problem
\item \verb{[10]} end of contract
\item \verb{[11]} seasonal work
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3877/vshow}{GSS Data Explorer page for \code{whyjbct1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab company's sales were declining \tab machines and computers made my job obsolete \tab job requirements changed and i was not able to do the job \tab job moved to another location in the united states \tab job moved to another country \tab other \tab Total \cr
   2008 \tab 1965 \tab 21 \tab 1 \tab 3 \tab 1 \tab 1 \tab 31 \tab 2023 \cr
   Total \tab 1965 \tab 21 \tab 1 \tab 3 \tab 1 \tab 1 \tab 31 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{curempyr}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{whyleave}},
\code{\link{work3yrs}}
}
\concept{Global Economics}
\keyword{variable}
