% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{whyjbend}
\alias{whyjbend}
\title{Main reason your job end}
\source{
General Social Survey https://gss.norc.org
}
\description{
whyjbend
}
\details{
Question What was the main reason that your job ended?
}
\section{Values}{

\itemize{
\item \code{1} i reached retirement age
\item \code{2} i retired early, by choice
\item \code{3} i retired early, not by choice
\item \code{4} i became (permanently) disabled
\item \code{5} my place of work was shut down
\item \code{6} i was dismissed
\item \code{7} my term of employment/contract ended
\item \code{8} family responsibilities
\item \code{9} i got married
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4595/vshow}{GSS Data Explorer page for \code{whyjbend}}.
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{worryjob}},
\code{\link{yrjobend}}

Other Unemployment: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{yrjobend}}
}
\concept{Unemployment}
\concept{Work Orientation}
\keyword{variable}
