% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{whypoor1}
\alias{whypoor1}
\title{Poor schools}
\source{
General Social Survey https://gss.norc.org
}
\description{
whypoor1
}
\details{
Question Now I will a list of reasons some people give to explain why there are poor people in this country. Please tell me whether you feel each of these is very important, somewhat important, or not important in explaining why there are poor people in this country.
A. Failure of society to provide good schools for many Americans
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} somewhat important
\item \code{3} not important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1170/vshow}{GSS Data Explorer page for \code{whypoor1}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthjews}},
\code{\link{wlthwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
