% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{whypoor4}
\alias{whypoor4}
\title{Lack of effort}
\source{
General Social Survey https://gss.norc.org
}
\description{
whypoor4
}
\details{
Question Now I will a list of reasons some people give to explain why there are poor people in this country. Please tell me whether you feel each of these is very important, somewhat important, or not important in explaining why there are poor people in this country.
D. Lack of effort by the poor themselves
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} somewhat important
\item \verb{[3]} not important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1173/vshow}{GSS Data Explorer page for \code{whypoor4}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very important \tab somewhat important \tab not important \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 620 \tab 602 \tab 115 \tab 29 \tab 6 \tab 1372 \cr
   Total \tab 0 \tab 620 \tab 602 \tab 115 \tab 29 \tab 6 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{wlthjews}},
\code{\link{wlthwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
