% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{whywkhme}
\alias{whywkhme}
\title{Usual reason r work at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
whywkhme
}
\details{
Question Is it usually because you want to, you have to in order to keep up with your job, or for some other reason?
}
\section{Values}{

\itemize{
\item \verb{[1]} worker wants to work at home
\item \verb{[2]} worker has to work at home to keep up with job
\item \verb{[3]} worker is operating home-based business
\item \verb{[4]} other combinations and other reasons
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2768/vshow}{GSS Data Explorer page for \code{whywkhme}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab worker wants to work at home \tab worker has to work at home to keep up with job \tab worker is operating home-based business \tab no answer \tab don't know \tab other combinations and other reasons \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 2077 \tab 254 \tab 224 \tab 197 \tab 13 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2006 \tab 3803 \tab 257 \tab 225 \tab 213 \tab 11 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2010 \tab 1531 \tab 160 \tab 189 \tab 60 \tab 27 \tab 3 \tab 74 \tab - \tab - \tab 2044 \cr
   2014 \tab 2033 \tab 142 \tab 203 \tab 84 \tab 7 \tab - \tab 69 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 171 \tab 200 \tab 76 \tab 896 \tab 3 \tab 102 \tab - \tab - \tab 2348 \cr
   2022 \tab 2641 \tab 476 \tab 160 \tab 106 \tab 3 \tab 8 \tab 130 \tab 20 \tab - \tab 3544 \cr
   Total \tab 12985 \tab 1460 \tab 1201 \tab 736 \tab 957 \tab 15 \tab 375 \tab 20 \tab 0 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{chngtme}},
\code{\link{condemnd}},
\code{\link{cowrkhlp}},
\code{\link{cowrkint}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{fairearn}},
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{fringeok}},
\code{\link{handmove}},
\code{\link{haveinfo}},
\code{\link{health1}},
\code{\link{height}},
\code{\link{hlpequip}},
\code{\link{hlthdays}},
\code{\link{hrsrelax}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{hyperten}},
\code{\link{jobfind1}},
\code{\link{jobsecok}},
\code{\link{knowschd}},
\code{\link{knowschd1}},
\code{\link{knowwhat}},
\code{\link{laidoff}},
\code{\link{learnnew}},
\code{\link{lotofsay}},
\code{\link{manvsemp}},
\code{\link{misswork}},
\code{\link{mntlhlth}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{myskills}},
\code{\link{ntwkhard}},
\code{\link{opdevel}},
\code{\link{overwork}},
\code{\link{painarms}},
\code{\link{partteam}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{prodctiv}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{respect}},
\code{\link{rincblls}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{satjob1}},
\code{\link{secondwk}},
\code{\link{setthngs}},
\code{\link{slpprblm}},
\code{\link{spvtrfair}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{suphelp}},
\code{\link{teamsafe}},
\code{\link{toofewwk}},
\code{\link{trainops}},
\code{\link{trdunion}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{usedup}},
\code{\link{usetech}},
\code{\link{waypaid}},
\code{\link{weight}},
\code{\link{wkageism}},
\code{\link{wkbonus}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkpraise}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksmooth}},
\code{\link{wktopsat}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrktime}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Employment: 
\code{\link{accntsci}},
\code{\link{advances}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compwage}},
\code{\link{curempyr}},
\code{\link{denyrais}},
\code{\link{dowell}},
\code{\link{drugtest}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empyears}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{evunemp}},
\code{\link{expernc}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findjob}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{futpromo}},
\code{\link{harjob5}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{ignorwk}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobyears}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{knowschd}},
\code{\link{laidoff}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leavejb}},
\code{\link{liedcwkr}},
\code{\link{losejb12}},
\code{\link{losejob5}},
\code{\link{lostjobs}},
\code{\link{ltrsref}},
\code{\link{marrcoun}},
\code{\link{medtreat}},
\code{\link{moredays}},
\code{\link{moretrde}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{neverwk}},
\code{\link{newjobs}},
\code{\link{numemp}},
\code{\link{numemply}},
\code{\link{numemps}},
\code{\link{numemps5}},
\code{\link{numemps_0422}},
\code{\link{numown}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{ownbiz}},
\code{\link{paidemps}},
\code{\link{partfull}},
\code{\link{paysame}},
\code{\link{physical}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{ryrsfirm}},
\code{\link{samejob}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{skiltest}},
\code{\link{slsmnshp}},
\code{\link{spwrkgvt}},
\code{\link{startbiz}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totunemp}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkbhvrs}},
\code{\link{work3yrs}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrkstat}},
\code{\link{wrktype}},
\code{\link{wrkyears}},
\code{\link{yearbiz}},
\code{\link{yearsjob}},
\code{\link{yrsfirm}}

Other Work: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{balwkfam}},
\code{\link{buyvalue}},
\code{\link{chngwork}},
\code{\link{cmprgmng}},
\code{\link{commute}},
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{denyrais}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{empinput}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{extr2005}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{getahead}},
\code{\link{getaheay}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{impjob}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jobcntrl}},
\code{\link{jobfind}},
\code{\link{jobfind1}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobpromo}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobskill}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{lottosay}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{monitred}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{numorg}},
\code{\link{offsup}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{parleave}},
\code{\link{pawrkslf}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{phyeffrt}},
\code{\link{placewrk}},
\code{\link{prideorg}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{richwork}},
\code{\link{rimpskls}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{satjob}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexhar}},
\code{\link{skltrain}},
\code{\link{slfmangd}},
\code{\link{slsmnshp}},
\code{\link{spdays}},
\code{\link{sphedslf}},
\code{\link{sphour}},
\code{\link{spvtrfair}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{spwrkslf}},
\code{\link{stayorg3}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{talkteam}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{toldwork}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{trynewjb}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{useskill}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkcontct}},
\code{\link{wkfambiz}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkracism}},
\code{\link{wkrspns}},
\code{\link{wksexism}},
\code{\link{wksub}},
\code{\link{wksubs}},
\code{\link{wksup}},
\code{\link{wksups}},
\code{\link{wktenure}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{workdy}},
\code{\link{workfor}},
\code{\link{workhr}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{worknow}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkhome}},
\code{\link{wrkimp}},
\code{\link{wrkindep}},
\code{\link{wrkindp}},
\code{\link{wrkmuch}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{wrkwell}},
\code{\link{yearsjob}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Employment}
\concept{Quality of Working Life}
\concept{Work}
\keyword{variable}
