% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{wirtap}
\alias{wirtap}
\title{Wiretapping}
\source{
General Social Survey https://gss.norc.org
}
\description{
wirtap
}
\details{
Question Everything considered, would you say that, in general, you approve or disapprove of wiretapping?
}
\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1974, 1975, 1977, 1978, 1982, 1983, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/284/vshow}{GSS Data Explorer page for \code{wirtap}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab approve \tab disapprove \tab don't know \tab no answer \tab Total \cr
   1974 \tab - \tab 245 \tab 1179 \tab 57 \tab 3 \tab 1484 \cr
   1975 \tab - \tab 242 \tab 1186 \tab 59 \tab 3 \tab 1490 \cr
   1977 \tab - \tab 282 \tab 1198 \tab 48 \tab 2 \tab 1530 \cr
   1978 \tab - \tab 291 \tab 1192 \tab 46 \tab 3 \tab 1532 \cr
   1982 \tab - \tab 311 \tab 1462 \tab 80 \tab 7 \tab 1860 \cr
   1983 \tab - \tab 296 \tab 1239 \tab 61 \tab 3 \tab 1599 \cr
   1985 \tab - \tab 349 \tab 1131 \tab 47 \tab 7 \tab 1534 \cr
   1986 \tab - \tab 326 \tab 1087 \tab 53 \tab 4 \tab 1470 \cr
   1988 \tab 493 \tab 200 \tab 727 \tab 56 \tab 5 \tab 1481 \cr
   1989 \tab 531 \tab 263 \tab 690 \tab 47 \tab 6 \tab 1537 \cr
   1990 \tab 444 \tab 203 \tab 668 \tab 54 \tab 3 \tab 1372 \cr
   1991 \tab 493 \tab 247 \tab 717 \tab 57 \tab 3 \tab 1517 \cr
   1993 \tab 526 \tab 245 \tab 768 \tab 61 \tab 6 \tab 1606 \cr
   1994 \tab 2485 \tab 93 \tab 385 \tab 27 \tab 2 \tab 2992 \cr
   Total \tab 4972 \tab 3593 \tab 13629 \tab 753 \tab 57 \tab 23004 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}}
}
\concept{Core}
\keyword{variable}
