% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{wkageism}
\alias{wkageism}
\title{R feels discriminated because of age}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkageism
}
\details{
Question Do you feel in any way discriminated against on your job because of your age?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2012, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2821/vshow}{GSS Data Explorer page for \code{wkageism}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 155 \tab 1620 \tab 7 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 148 \tab 1565 \tab 4 \tab 17 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 89 \tab 1073 \tab 2 \tab 23 \tab - \tab 2044 \cr
   2012 \tab 795 \tab 108 \tab 1040 \tab 4 \tab 27 \tab - \tab 1974 \cr
   2014 \tab 1288 \tab 78 \tab 1166 \tab - \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 125 \tab 1287 \tab 4 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 155 \tab 1799 \tab 6 \tab 13 \tab 16 \tab 3544 \cr
   Total \tab 9140 \tab 858 \tab 9550 \tab 27 \tab 132 \tab 16 \tab 19723 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}},
\code{\link{wksexism}}

Other Workplace Conflict: 
\code{\link{health1}},
\code{\link{trynewjb}},
\code{\link{wkracism}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
