% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{wkbhvrs}
\alias{wkbhvrs}
\title{Employee doing these things at work to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkbhvrs
}
\details{
Question In the past 12 months, the person who has engaged in these types of behaviors has most often been:
}
\section{Values}{

\itemize{
\item \verb{[1]} a co-worker
\item \verb{[2]} your supervisor or boss
\item \verb{[3]} a customer/client/patient
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3022/vshow}{GSS Data Explorer page for \code{wkbhvrs}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a co-worker \tab your supervisor or boss \tab a customer/client/patient \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1219 \tab 324 \tab 175 \tab 180 \tab 52 \tab 24 \tab 1974 \cr
   Total \tab 1219 \tab 324 \tab 175 \tab 180 \tab 52 \tab 24 \tab 1974 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{denyrais}},
\code{\link{eharaswk}},
\code{\link{ignorwk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{wkrspns}}
}
\concept{Workplace Conflict}
\keyword{variable}
