% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{wkbonus}
\alias{wkbonus}
\title{R is likely to get a bonus or pay increase}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkbonus
}
\details{
Question When you do your job well, are you likely to get a bonus or pay increase?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} maybe
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2815/vshow}{GSS Data Explorer page for \code{wkbonus}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab maybe \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 431 \tab 251 \tab 1082 \tab 12 \tab 20 \tab 2765 \cr
   2006 \tab 2776 \tab 400 \tab 288 \tab 996 \tab 26 \tab 24 \tab 4510 \cr
   Total \tab 3745 \tab 831 \tab 539 \tab 2078 \tab 38 \tab 44 \tab 7275 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
