% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{wkfambiz}
\alias{wkfambiz}
\title{Work for pay in a farm etc owned by your family?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkfambiz
}
\details{
Question Do you work for pay in a farm or other business owned in whole or in substantial part by a member or members of your family (parents, grandparents, children, brothers, sisters, aunts, uncles, or first cousins)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1876/vshow}{GSS Data Explorer page for \code{wkfambiz}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1996 \tab 1460 \tab 54 \tab 1381 \tab 9 \tab - \tab 2904 \cr
   Total \tab 1460 \tab 54 \tab 1381 \tab 9 \tab 0 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{borrowed}},
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{econsys}},
\code{\link{evbuyhme}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{fixer1}},
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{fixhome}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{haggle}},
\code{\link{homeyear}},
\code{\link{hrfambiz}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal3}},
\code{\link{inequal5}},
\code{\link{lawyer}},
\code{\link{lawyer1}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{layoffs}},
\code{\link{learnhme}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{natrecon}},
\code{\link{newowned}},
\code{\link{newused}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{realtor1}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{reldealr}},
\code{\link{relfixer}},
\code{\link{relhome}},
\code{\link{rellaw}},
\code{\link{relprivt}},
\code{\link{reqinfo}},
\code{\link{satcar}},
\code{\link{satfixes}},
\code{\link{sathome}},
\code{\link{satlawyr}},
\code{\link{sellauto}},
\code{\link{sellbaby}},
\code{\link{sellbed}},
\code{\link{sellhome}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{tagsales}},
\code{\link{transoth}},
\code{\link{transrel}},
\code{\link{typdealr}},
\code{\link{warrntyc}},
\code{\link{warrntyh}},
\code{\link{wholoand}},
\code{\link{whosold}}

Other Agriculture: 
\code{\link{actfarm}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{fambiz}},
\code{\link{givfarm}},
\code{\link{intfarm}},
\code{\link{memfarm}},
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{numfarm}},
\code{\link{payfarm}},
\code{\link{solfarm}},
\code{\link{typfarm}},
\code{\link{yrfarm2}},
\code{\link{yrfarm3}}

Other Labor Force Status: 
\code{\link{evwork}},
\code{\link{hrfambiz}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{kdevwork}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{mawk16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawork14}},
\code{\link{mawrkgrw}},
\code{\link{neverwk}},
\code{\link{partfull}},
\code{\link{spevwork}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spwrksta}},
\code{\link{unemp}},
\code{\link{weekswrk}},
\code{\link{wrkstat}}

Other Occupation: 
\code{\link{accntsci}},
\code{\link{ageemp}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{cmprgmng}},
\code{\link{coisco08}},
\code{\link{commute}},
\code{\link{copres105plus}},
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{cosei10inc}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{datefrst}},
\code{\link{dowell}},
\code{\link{educemp}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{expernc}},
\code{\link{fambiz}},
\code{\link{famresp}},
\code{\link{farming}},
\code{\link{findnwjb}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstsei}},
\code{\link{firstslf}},
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givexec}},
\code{\link{givfarm}},
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givlaw}},
\code{\link{givmason}},
\code{\link{givowner}},
\code{\link{givrocc}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{indfirst}},
\code{\link{indus10}},
\code{\link{indus80}},
\code{\link{industry}},
\code{\link{isco08}},
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{isco88}},
\code{\link{journlsm}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{maind10}},
\code{\link{maind80}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{maisco88}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{marrcoun}},
\code{\link{masei}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{monthly}},
\code{\link{occ}},
\code{\link{occ10}},
\code{\link{occ80}},
\code{\link{occfirst}},
\code{\link{occmobil}},
\code{\link{ownbiz}},
\code{\link{paind10}},
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paisc681}},
\code{\link{paisco08}},
\code{\link{paisco68}},
\code{\link{paisco88}},
\code{\link{pajob}},
\code{\link{paocc10}},
\code{\link{paocc16}},
\code{\link{paocc80}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pasei}},
\code{\link{pasei10}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{paslf}},
\code{\link{pastpay}},
\code{\link{pasup}},
\code{\link{pawrkslf}},
\code{\link{paybus}},
\code{\link{paycabnt}},
\code{\link{paychnge}},
\code{\link{payclerk}},
\code{\link{paydoc}},
\code{\link{payexec}},
\code{\link{payfarm}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{paymason}},
\code{\link{payowner}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{paysame}},
\code{\link{paysec}},
\code{\link{payskill}},
\code{\link{payunskl}},
\code{\link{presfrst}},
\code{\link{prestg80}},
\code{\link{prestige}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{ryrsfirm}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{sei}},
\code{\link{sei10}},
\code{\link{sei10educ}},
\code{\link{sei10inc}},
\code{\link{sexemp}},
\code{\link{slsmnshp}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{sphedind}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{sphedslf}},
\code{\link{spind}},
\code{\link{spind10}},
\code{\link{spind80}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco68}},
\code{\link{spisco88}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spsei}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}},
\code{\link{spwrkslf}},
\code{\link{weekly}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{yearly}},
\code{\link{yrsfirm}}

Other Work: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{balwkfam}},
\code{\link{buyvalue}},
\code{\link{chngwork}},
\code{\link{cmprgmng}},
\code{\link{commute}},
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{denyrais}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{empinput}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{extr2005}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{getahead}},
\code{\link{getaheay}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{impjob}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jobcntrl}},
\code{\link{jobfind}},
\code{\link{jobfind1}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobpromo}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobskill}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{lottosay}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{monitred}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{numorg}},
\code{\link{offsup}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{parleave}},
\code{\link{pawrkslf}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{phyeffrt}},
\code{\link{placewrk}},
\code{\link{prideorg}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{richwork}},
\code{\link{rimpskls}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{satjob}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexhar}},
\code{\link{skltrain}},
\code{\link{slfmangd}},
\code{\link{slsmnshp}},
\code{\link{spdays}},
\code{\link{sphedslf}},
\code{\link{sphour}},
\code{\link{spvtrfair}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{spwrkslf}},
\code{\link{stayorg3}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{talkteam}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{toldwork}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{trynewjb}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{useskill}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkcontct}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkracism}},
\code{\link{wkrspns}},
\code{\link{wksexism}},
\code{\link{wksub}},
\code{\link{wksubs}},
\code{\link{wksup}},
\code{\link{wksups}},
\code{\link{wktenure}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{workdy}},
\code{\link{workfor}},
\code{\link{workhr}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{worknow}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkhome}},
\code{\link{wrkimp}},
\code{\link{wrkindep}},
\code{\link{wrkindp}},
\code{\link{wrkmuch}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{wrkwell}},
\code{\link{yearsjob}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Agriculture}
\concept{Labor Force Status}
\concept{Markets}
\concept{Occupation}
\concept{Work}
\keyword{variable}
