% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{wkharoth}
\alias{wkharoth}
\title{R threatened on the job last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkharoth
}
\details{
Question In the last 12 months, were you threatened or harassed in any other way by anyone while you were on the job?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2825/vshow}{GSS Data Explorer page for \code{wkharoth}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 202 \tab 1572 \tab 6 \tab 16 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 151 \tab 1562 \tab 5 \tab 16 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 111 \tab 1050 \tab 2 \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 101 \tab 1143 \tab 1 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 120 \tab 1290 \tab 4 \tab 34 \tab - \tab 2348 \cr
   2022 \tab 1556 \tab 192 \tab 1764 \tab 2 \tab 13 \tab 17 \tab 3544 \cr
   Total \tab 8346 \tab 877 \tab 8381 \tab 20 \tab 108 \tab 17 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkharsex}},
\code{\link{wkracism}},
\code{\link{wksexism}}
}
\concept{Quality of Working Life}
\keyword{variable}
