% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{wkkidscl}
\alias{wkkidscl}
\title{Did r work outside home with child under school age}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkkidscl
}
\details{
Question Did you work outside the home full-time, part-time, or not at all…
A. When a child was under school age?
}
\section{Values}{

\itemize{
\item \code{1} worked full-time
\item \code{2} worked part-time
\item \code{3} stayed at home
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4479/vshow}{GSS Data Explorer page for \code{wkkidscl}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab worked full-time \tab worked part-time \tab stayed at home \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1044 \tab 615 \tab 121 \tab 147 \tab 39 \tab 8 \tab 1974 \cr
   Total \tab 1044 \tab 615 \tab 121 \tab 147 \tab 39 \tab 8 \tab 1974 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
