% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{wkkidscs}
\alias{wkkidscs}
\title{Did partner work outside home with child under school age}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkkidscs
}
\details{
Question What about your (spouse/partner) at that time – did (he/she) work outside the home full-time, part-time, or not at all…
A. When a child was under school age?
}
\section{Values}{

\itemize{
\item \code{1} worked full-time
\item \code{2} worked part-time
\item \code{3} stayed at home
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4481/vshow}{GSS Data Explorer page for \code{wkkidscs}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}

Other Work: 
\code{\link{famvswk1}},
\code{\link{jobvsfa1}},
\code{\link{wkkidscl}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\concept{Work}
\keyword{variable}
