% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{wkpersnl}
\alias{wkpersnl}
\title{Importance of personal contact at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkpersnl
}
\details{
Question (HAND CARD D2)
(For each of the following, please tell me how important you personally think it is in a job. How important is...)
...a job that involves personal contact with other people?
(Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} neither important nor unimportant
\item \code{4} not important
\item \code{5} not important at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6322/vshow}{GSS Data Explorer page for \code{wkpersnl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab neither important nor unimportant \tab not important \tab not important at all \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 323 \tab 628 \tab 354 \tab 140 \tab 23 \tab 3 \tab 6 \tab 2867 \cr
   Total \tab 1390 \tab 323 \tab 628 \tab 354 \tab 140 \tab 23 \tab 3 \tab 6 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
