% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{wkracism}
\alias{wkracism}
\title{R feels discriminated because of race}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkracism
}
\details{
Question Do you feel in any way discriminated against on your job because of your race or ethnic origin?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2012, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2822/vshow}{GSS Data Explorer page for \code{wkracism}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 98 \tab 1676 \tab 8 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 82 \tab 1631 \tab 4 \tab 17 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 66 \tab 1096 \tab 2 \tab 23 \tab - \tab 2044 \cr
   2012 \tab 795 \tab 80 \tab 1068 \tab 4 \tab 27 \tab - \tab 1974 \cr
   2014 \tab 1288 \tab 47 \tab 1195 \tab 1 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 80 \tab 1330 \tab 6 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 105 \tab 1845 \tab 7 \tab 14 \tab 18 \tab 3544 \cr
   Total \tab 9140 \tab 558 \tab 9841 \tab 32 \tab 134 \tab 18 \tab 19723 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wksexism}}

Other Workplace Conflict: 
\code{\link{health1}},
\code{\link{trynewjb}},
\code{\link{wkageism}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
