% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{wkrspns}
\alias{wkrspns}
\title{Response to misbehavior from coworkers}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkrspns
}
\details{
Question What did you typically do in response to these behaviors?
}
\section{Values}{

\itemize{
\item \verb{[1]} did nothing (ignored it)
\item \verb{[2]} talked to co-workers or other people at work
\item \verb{[3]} talked to family or friends
\item \verb{[4]} reported the incident to management or those higher up
\item \verb{[5]} reported the incident to the union
\item \verb{[6]} confronted the person (s) engaging in the behaviors
\item \verb{[7]} filed a formal complaint or grievance about the behavior or experiences
\item \verb{[8]} did not know what to do
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3023/vshow}{GSS Data Explorer page for \code{wkrspns}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab did nothing (ignored it) \tab talked to co-workers or other people at work \tab talked to family or friends \tab reported the incident to management or those higher up \tab reported the incident to the union \tab confronted the person (s) engaging in the behaviors \tab filed a formal complaint or grievance about the behavior or experiences \tab did not know what to do \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2012 \tab 1219 \tab 214 \tab 136 \tab 59 \tab 109 \tab 6 \tab 185 \tab 9 \tab 14 \tab 15 \tab 8 \tab - \tab 1974 \cr
   Total \tab 1219 \tab 214 \tab 136 \tab 59 \tab 109 \tab 6 \tab 185 \tab 9 \tab 14 \tab 15 \tab 8 \tab 0 \tab 1974 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{denyrais}},
\code{\link{difstand}},
\code{\link{eharaswk}},
\code{\link{gdjobsec}},
\code{\link{health1}},
\code{\link{hotargus}},
\code{\link{ignorwk}},
\code{\link{jobsecok}},
\code{\link{jokeswk}},
\code{\link{lackinfo}},
\code{\link{liedcwkr}},
\code{\link{lookaway}},
\code{\link{manvsemp}},
\code{\link{mntlhlth}},
\code{\link{numemps}},
\code{\link{numemps_0422}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{physharm}},
\code{\link{putdown}},
\code{\link{reliedon}},
\code{\link{rptprobs}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{supcares}},
\code{\link{treatres}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkracism}},
\code{\link{wkvsfam}}

Other Work: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{balwkfam}},
\code{\link{buyvalue}},
\code{\link{chngwork}},
\code{\link{cmprgmng}},
\code{\link{commute}},
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{denyrais}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{empinput}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{extr2005}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{getahead}},
\code{\link{getaheay}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{impjob}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jobcntrl}},
\code{\link{jobfind}},
\code{\link{jobfind1}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobpromo}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobskill}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{lottosay}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{monitred}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{numorg}},
\code{\link{offsup}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{parleave}},
\code{\link{pawrkslf}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{phyeffrt}},
\code{\link{placewrk}},
\code{\link{prideorg}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{richwork}},
\code{\link{rimpskls}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{satjob}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexhar}},
\code{\link{skltrain}},
\code{\link{slfmangd}},
\code{\link{slsmnshp}},
\code{\link{spdays}},
\code{\link{sphedslf}},
\code{\link{sphour}},
\code{\link{spvtrfair}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{spwrkslf}},
\code{\link{stayorg3}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{talkteam}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{toldwork}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{trynewjb}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{useskill}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkcontct}},
\code{\link{wkfambiz}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksub}},
\code{\link{wksubs}},
\code{\link{wksup}},
\code{\link{wksups}},
\code{\link{wktenure}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{workdy}},
\code{\link{workfor}},
\code{\link{workhr}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{worknow}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkhome}},
\code{\link{wrkimp}},
\code{\link{wrkindep}},
\code{\link{wrkindp}},
\code{\link{wrkmuch}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{wrkwell}},
\code{\link{yearsjob}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Work}
\concept{Workplace Conflict}
\keyword{variable}
