% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{wksexism}
\alias{wksexism}
\title{R feels discriminated because of gender}
\source{
General Social Survey https://gss.norc.org
}
\description{
wksexism
}
\details{
Question Do you feel in any way discriminated against on your job because of your gender?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2823/vshow}{GSS Data Explorer page for \code{wksexism}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 131 \tab 1644 \tab 7 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 119 \tab 1594 \tab 5 \tab 16 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 58 \tab 1103 \tab 3 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 57 \tab 1187 \tab - \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 95 \tab 1315 \tab 6 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 137 \tab 1811 \tab 9 \tab 14 \tab 18 \tab 3544 \cr
   Total \tab 8345 \tab 597 \tab 8654 \tab 30 \tab 105 \tab 18 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}}
}
\concept{Quality of Working Life}
\keyword{variable}
