% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_64.R
\name{wksub}
\alias{wksub}
\title{Does r or spouse have supervisor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wksub
}
\details{
Question Do you (does your (SPOUSE)) have a supervisor on (your/his/her) job to whom you are (he/she/is) directly responsible?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/C/-/- \tab partial \cr
   1989 \tab A/C/-/- \tab partial \cr
   1990 \tab A/C/-/- \tab partial \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/C/-/- \tab partial \cr
   1994 \tab A/C/-/- \tab partial \cr
   1996 \tab A/C/-/- \tab partial \cr
   1998 \tab A/C/-/- \tab partial \cr
   2000 \tab A/C/-/- \tab partial \cr
   2002 \tab A/C/-/- \tab partial \cr
   2004 \tab A/C/-/- \tab partial \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/577/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 416 \tab 241 \tab 21 \tab 935 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 369 \tab 242 \tab 31 \tab 858 \tab 4 \tab - \tab - \tab 1504 \cr
   1974 \tab 400 \tab 227 \tab 20 \tab 836 \tab 1 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 450 \tab 194 \tab 31 \tab 822 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab 382 \tab 194 \tab 26 \tab 925 \tab 3 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 404 \tab 212 \tab 4 \tab 848 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 589 \tab 241 \tab 44 \tab 983 \tab 3 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 390 \tab 212 \tab 4 \tab 864 \tab 3 \tab - \tab - \tab 1473 \cr
   1985 \tab 438 \tab 251 \tab 7 \tab 838 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 506 \tab 264 \tab 12 \tab 1033 \tab 4 \tab - \tab - \tab 1819 \cr
   1988 \tab 792 \tab 109 \tab 91 \tab 488 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 808 \tab 147 \tab 7 \tab 572 \tab 3 \tab - \tab - \tab 1537 \cr
   1990 \tab 723 \tab 125 \tab 4 \tab 518 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 527 \tab 200 \tab 13 \tab 776 \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab 839 \tab 144 \tab 3 \tab 620 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1532 \tab 278 \tab 7 \tab 1174 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 1420 \tab 302 \tab 10 \tab 1170 \tab 2 \tab - \tab - \tab 2904 \cr
   1998 \tab 1435 \tab 250 \tab 22 \tab 1122 \tab 3 \tab - \tab - \tab 2832 \cr
   2000 \tab 1467 \tab 213 \tab 38 \tab 1096 \tab 3 \tab - \tab - \tab 2817 \cr
   2002 \tab 1441 \tab 264 \tab 2 \tab 1057 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 1397 \tab 290 \tab 22 \tab 1100 \tab 3 \tab - \tab - \tab 2812 \cr
   2006 \tab 1289 \tab 615 \tab 16 \tab 2585 \tab 5 \tab - \tab - \tab 4510 \cr
   2008 \tab 581 \tab 281 \tab 6 \tab 1149 \tab 6 \tab - \tab - \tab 2023 \cr
   2010 \tab 700 \tab 264 \tab 15 \tab 1063 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 607 \tab 229 \tab 11 \tab 1122 \tab 5 \tab - \tab - \tab 1974 \cr
   2014 \tab 788 \tab 316 \tab 3 \tab 1427 \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab 941 \tab 348 \tab 16 \tab 1558 \tab 4 \tab - \tab - \tab 2867 \cr
   2018 \tab 734 \tab 282 \tab 2 \tab 1325 \tab 5 \tab - \tab - \tab 2348 \cr
   2021 \tab 1471 \tab 423 \tab 17 \tab 2077 \tab 38 \tab 6 \tab - \tab 4032 \cr
   2022 \tab 1267 \tab 390 \tab 6 \tab 1855 \tab 21 \tab 5 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 31194 \tab 7748 \tab 511 \tab 32796 \tab 130 \tab 11 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{incneed}},
\code{\link{mininc}},
\code{\link{satfin}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}}

Other Job: 
\code{\link{wksup}}

Other Work: 
\code{\link{wksubs}},
\code{\link{wksup}}
}
\concept{Core}
\concept{Job}
\concept{Work}
\keyword{variable}
