% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{wksup1}
\alias{wksup1}
\title{Does r or spouse or partner supervise anyone}
\source{
General Social Survey https://gss.norc.org
}
\description{
wksup1
}
\details{
Question In your (SPOUSE or PARTNER'S) job, (do you/does he/she) supervise anyone who is directly responsible to (you/him/her)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6178/vshow}{GSS Data Explorer page for \code{wksup1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2016 \tab 914 \tab 749 \tab 1197 \tab 5 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 693 \tab 595 \tab 1055 \tab 4 \tab 1 \tab - \tab 2348 \cr
   2021 \tab 1419 \tab 844 \tab 1704 \tab 59 \tab 1 \tab 5 \tab 4032 \cr
   2022 \tab 1203 \tab 788 \tab 1527 \tab 19 \tab 5 \tab 2 \tab 3544 \cr
   2024 \tab 1250 \tab 605 \tab 1406 \tab 30 \tab 10 \tab 8 \tab 3309 \cr
   Total \tab 5479 \tab 3581 \tab 6889 \tab 117 \tab 19 \tab 15 \tab 16100 \cr
}
}

\seealso{
Other Core: 
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{parsol}},
\code{\link{unemp}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{wksups}},
\code{\link{wksups1}}
}
\concept{Core}
\keyword{variable}
