% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{wktopsat}
\alias{wktopsat}
\title{Satisfaction comes from work}
\source{
General Social Survey https://gss.norc.org
}
\description{
wktopsat
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
E. My main satisfaction in life comes from work
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2778/vshow}{GSS Data Explorer page for \code{wktopsat}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 128 \tab 369 \tab 883 \tab 389 \tab 14 \tab 13 \tab 2765 \cr
   2006 \tab 2776 \tab 99 \tab 387 \tab 832 \tab 393 \tab 12 \tab 11 \tab 4510 \cr
   Total \tab 3745 \tab 227 \tab 756 \tab 1715 \tab 782 \tab 26 \tab 24 \tab 7275 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{respect}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{trustman}}
}
\concept{Quality of Working Life}
\keyword{variable}
