% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{wkwkends}
\alias{wkwkends}
\title{Do you work weekends}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkwkends
}
\details{
Question (HAND CARD D8)
And how often...
...does your job involve working on weekends?
(Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} hardly ever
\item \verb{[5]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6331/vshow}{GSS Data Explorer page for \code{wkwkends}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 198 \tab 177 \tab 234 \tab 146 \tab 180 \tab 5 \tab 2867 \cr
   Total \tab 1927 \tab 198 \tab 177 \tab 234 \tab 146 \tab 180 \tab 5 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
