% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{wlthasns}
\alias{wlthasns}
\title{Rich - poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthasns
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
4. Asian Americans?
}
\section{Values}{

\itemize{
\item \code{1} 1 - rich
\item \code{7} 7 - poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1181/vshow}{GSS Data Explorer page for \code{wlthasns}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 - rich \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - poor \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 17 \tab 73 \tab 214 \tab 437 \tab 286 \tab 143 \tab 57 \tab 128 \tab 17 \tab 1372 \cr
   1994 \tab 1518 \tab 40 \tab 95 \tab 244 \tab 484 \tab 277 \tab 147 \tab 42 \tab 132 \tab 13 \tab 2992 \cr
   2000 \tab 1419 \tab 41 \tab 143 \tab 270 \tab 476 \tab 198 \tab 86 \tab 24 \tab 146 \tab 14 \tab 2817 \cr
   Total \tab 2937 \tab 98 \tab 311 \tab 728 \tab 1397 \tab 761 \tab 376 \tab 123 \tab 406 \tab 44 \tab 7181 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthso}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
