% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{wlthblks}
\alias{wlthblks}
\title{Rich - poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthblks
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
3. Blacks?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - rich
\item \verb{[7]} 7 - poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1996, 1998, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1990, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1180/vshow}{GSS Data Explorer page for \code{wlthblks}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - rich \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 3 \tab 19 \tab 40 \tab 275 \tab 491 \tab 365 \tab 108 \tab 54 \tab 17 \tab - \tab 1372 \cr
   1994 \tab 1518 \tab 5 \tab 18 \tab 46 \tab 298 \tab 603 \tab 360 \tab 92 \tab 39 \tab 13 \tab - \tab 2992 \cr
   1996 \tab 1935 \tab 3 \tab 24 \tab 60 \tab 220 \tab 363 \tab 216 \tab 40 \tab 38 \tab 5 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 12 \tab 37 \tab 123 \tab 436 \tab 721 \tab 402 \tab 54 \tab 77 \tab 9 \tab - \tab 2832 \cr
   2000 \tab 486 \tab 7 \tab 32 \tab 145 \tab 593 \tab 862 \tab 480 \tab 112 \tab 76 \tab 24 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 7 \tab 8 \tab 59 \tab 250 \tab 344 \tab 183 \tab 38 \tab 16 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 1 \tab 9 \tab 59 \tab 235 \tab 383 \tab 171 \tab 27 \tab 18 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 6 \tab 21 \tab 111 \tab 498 \tab 799 \tab 407 \tab 93 \tab 48 \tab 9 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 2 \tab 27 \tab 83 \tab 362 \tab 543 \tab 223 \tab 54 \tab 28 \tab 7 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 9 \tab 16 \tab 88 \tab 423 \tab 549 \tab 240 \tab 59 \tab 35 \tab 11 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 2 \tab 22 \tab 76 \tab 385 \tab 523 \tab 222 \tab 40 \tab 23 \tab 9 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 6 \tab 27 \tab 99 \tab 515 \tab 657 \tab 283 \tab 54 \tab 27 \tab 7 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 9 \tab 20 \tab 95 \tab 627 \tab 712 \tab 332 \tab 61 \tab 22 \tab 10 \tab - \tab 2867 \cr
   2018 \tab 1170 \tab 3 \tab 12 \tab 64 \tab 385 \tab 456 \tab 193 \tab 41 \tab 16 \tab 8 \tab - \tab 2348 \cr
   2021 \tab 2689 \tab 6 \tab 9 \tab 59 \tab 404 \tab 559 \tab 233 \tab 64 \tab 2 \tab 1 \tab 6 \tab 4032 \cr
   2022 \tab 2339 \tab 6 \tab 21 \tab 46 \tab 427 \tab 446 \tab 187 \tab 45 \tab 17 \tab 9 \tab 1 \tab 3544 \cr
   2024 \tab 2200 \tab 7 \tab 10 \tab 60 \tab 369 \tab 421 \tab 147 \tab 54 \tab 27 \tab 6 \tab 8 \tab 3309 \cr
   Total \tab 23401 \tab 94 \tab 332 \tab 1313 \tab 6702 \tab 9432 \tab 4644 \tab 1036 \tab 563 \tab 151 \tab 15 \tab 47683 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Blacks: 
\code{\link{acqblack}},
\code{\link{acqfmblk}},
\code{\link{acqnhblk}},
\code{\link{acqvablk}},
\code{\link{acqwkblk}},
\code{\link{admirblk}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkgovt}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blksimp}},
\code{\link{blkwrk}},
\code{\link{closeblk}},
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{comblk}},
\code{\link{discaff}},
\code{\link{discaffy}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fairblks}},
\code{\link{famblks}},
\code{\link{fareblks}},
\code{\link{feelblks}},
\code{\link{influblk}},
\code{\link{intlblks}},
\code{\link{jobaff}},
\code{\link{knwblk}},
\code{\link{liveblks}},
\code{\link{marblk}},
\code{\link{patrblks}},
\code{\link{symptblk}},
\code{\link{trtblack}},
\code{\link{usblk}},
\code{\link{violblks}},
\code{\link{workblks}},
\code{\link{wrkwayup}}
}
\concept{Blacks}
\concept{Intergroup Relations}
\keyword{variable}
