% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{wlthhsps}
\alias{wlthhsps}
\title{Rich - poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthhsps
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
5. Hispanic Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - rich
\item \verb{[7]} 7 - poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1990, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1182/vshow}{GSS Data Explorer page for \code{wlthhsps}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - rich \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 2 \tab 25 \tab 57 \tab 211 \tab 476 \tab 359 \tab 112 \tab 113 \tab 17 \tab - \tab 1372 \cr
   1994 \tab 1518 \tab 4 \tab 18 \tab 53 \tab 281 \tab 500 \tab 399 \tab 102 \tab 104 \tab 13 \tab - \tab 2992 \cr
   2000 \tab 1419 \tab 8 \tab 27 \tab 80 \tab 303 \tab 461 \tab 326 \tab 75 \tab 104 \tab 14 \tab - \tab 2817 \cr
   2018 \tab 1554 \tab 4 \tab 10 \tab 49 \tab 273 \tab 297 \tab 118 \tab 22 \tab 15 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 2689 \tab 3 \tab 18 \tab 66 \tab 442 \tab 533 \tab 192 \tab 74 \tab 4 \tab 1 \tab 10 \tab 4032 \cr
   2022 \tab 2339 \tab 11 \tab 25 \tab 70 \tab 431 \tab 421 \tab 171 \tab 40 \tab 28 \tab 8 \tab - \tab 3544 \cr
   Total \tab 9519 \tab 32 \tab 123 \tab 375 \tab 1941 \tab 2688 \tab 1565 \tab 425 \tab 368 \tab 59 \tab 10 \tab 17105 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthso}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
