% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{wlthjews}
\alias{wlthjews}
\title{Rich - poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthjews
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
2. Jews?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - rich
\item \verb{[7]} 7 - poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1179/vshow}{GSS Data Explorer page for \code{wlthjews}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 - rich \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - poor \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 152 \tab 336 \tab 350 \tab 323 \tab 68 \tab 15 \tab 6 \tab 105 \tab 17 \tab 1372 \cr
   2000 \tab 1419 \tab 138 \tab 283 \tab 325 \tab 328 \tab 73 \tab 33 \tab 14 \tab 188 \tab 16 \tab 2817 \cr
   Total \tab 1419 \tab 290 \tab 619 \tab 675 \tab 651 \tab 141 \tab 48 \tab 20 \tab 293 \tab 33 \tab 4189 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
