% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{wocomptr}
\alias{wocomptr}
\title{R can do job without computer}
\source{
General Social Survey https://gss.norc.org
}
\description{
wocomptr
}
\details{
Question Could you do your job without a computer or computerized equipment? Would you say….
}
\section{Values}{

\itemize{
\item \verb{[1]} couldn't do my job without a computer or computerized equipment
\item \verb{[2]} could do my job without a computer or computerized equipment but it would be much harder
\item \verb{[3]} doing my job without a computer or computerized equipment would be about the same as now
\item \verb{[4]} already do most of my job without a computer or computerized equipment
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3888/vshow}{GSS Data Explorer page for \code{wocomptr}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab couldn't do my job without a computer or computerized equipment \tab could do my job without a computer or computerized equipment but it would be much harder \tab doing my job without a computer or computerized equipment would be about the same as now \tab already do most of my job without a computer or computerized equipment \tab not available in this year \tab Total \cr
   2008 \tab 1381 \tab 289 \tab 248 \tab 52 \tab 53 \tab - \tab 2023 \cr
   Total \tab 1381 \tab 289 \tab 248 \tab 52 \tab 53 \tab 0 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{autonojb}},
\code{\link{bornsp}},
\code{\link{curempyr}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{natdeg}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{typevisa}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{work3yrs}}

Other Computers And Internet: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{autonojb}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{intaccss}},
\code{\link{internet}},
\code{\link{intrhome}},
\code{\link{invest12}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numhome}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polinf12}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{poltough}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{wkcomptr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhr}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}

Other Job: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{actupset}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{baseofop}},
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{bosswrks}},
\code{\link{chngeorg}},
\code{\link{chngtme}},
\code{\link{cmprgmng}},
\code{\link{compperf}},
\code{\link{condemnd}},
\code{\link{cowrkers}},
\code{\link{cowrkhlp}},
\code{\link{cowrkhrd}},
\code{\link{cowrkint}},
\code{\link{cowrkrel}},
\code{\link{curempyr}},
\code{\link{daily}},
\code{\link{dailywrk}},
\code{\link{danger}},
\code{\link{datefrst}},
\code{\link{denyrais}},
\code{\link{deptperf}},
\code{\link{difstand}},
\code{\link{discaff}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{discaffy}},
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{donothng}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{emphlth}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{emphplan}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{exptrain}},
\code{\link{extr2001}},
\code{\link{extr2005}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{famintjb}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{fehire}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{findnwjb}},
\code{\link{findout}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{formltrn}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{fringeok}},
\code{\link{ftwork}},
\code{\link{ftwork1}},
\code{\link{gdjobsec}},
\code{\link{genderwk}},
\code{\link{handmove}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{haveinfo}},
\code{\link{helporg}},
\code{\link{hiinc}},
\code{\link{hlpequip}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hlthtype}},
\code{\link{hotargus}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{hrsmoney}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{imatter}},
\code{\link{inclstjb}},
\code{\link{indperf}},
\code{\link{indusoth}},
\code{\link{intjob}},
\code{\link{intlcowk}},
\code{\link{iwrkhard}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobaff}},
\code{\link{jobcntct}},
\code{\link{jobcntrl}},
\code{\link{jobcowrk}},
\code{\link{jobeasy}},
\code{\link{jobfrnds}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobinfo}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobpay}},
\code{\link{jobplace}},
\code{\link{jobpromo}},
\code{\link{jobpub}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobsecok}},
\code{\link{jobskill}},
\code{\link{jobtvrad}},
\code{\link{jobvshme}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{knowschd}},
\code{\link{knowwhat}},
\code{\link{lackinfo}},
\code{\link{lastjob}},
\code{\link{lastyrwk}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{learning}},
\code{\link{learnjob}},
\code{\link{learnnew}},
\code{\link{leavejb}},
\code{\link{leavejob}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{localnum}},
\code{\link{locyrs}},
\code{\link{lookaway}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{lotofsay}},
\code{\link{lottosay}},
\code{\link{manvsemp}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{misswork}},
\code{\link{monitred}},
\code{\link{monthly}},
\code{\link{moonlite}},
\code{\link{moretrde}},
\code{\link{mykids}},
\code{\link{myskills}},
\code{\link{negjob5}},
\code{\link{newjob}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{ntwkhard}},
\code{\link{numemp}},
\code{\link{numemps5}},
\code{\link{numorg}},
\code{\link{numown}},
\code{\link{numsites}},
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{offsup}},
\code{\link{opdevel}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{orgsize}},
\code{\link{othcredt}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{otherwrk}},
\code{\link{othhrs}},
\code{\link{othplan}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{overwork}},
\code{\link{owninc}},
\code{\link{ownstock}},
\code{\link{paidemps}},
\code{\link{parleave}},
\code{\link{partorg}},
\code{\link{partteam}},
\code{\link{pastpay}},
\code{\link{pawrkslf}},
\code{\link{paychnge}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{perspace}},
\code{\link{phyeffrt}},
\code{\link{physharm}},
\code{\link{physwrk}},
\code{\link{placewrk}},
\code{\link{planwrk}},
\code{\link{privgovt}},
\code{\link{prodctiv}},
\code{\link{promotn}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{proudorg}},
\code{\link{putdown}},
\code{\link{rageemp}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{rdiscaff}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{reliedon}},
\code{\link{replaceu}},
\code{\link{respect}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rimmdisc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpaysame}},
\code{\link{rpromotn}},
\code{\link{rptprobs}},
\code{\link{rsecjob}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{rwrkindp}},
\code{\link{ryrsfirm}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{samevals}},
\code{\link{samework}},
\code{\link{satjob}},
\code{\link{satjob1}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{schoolng}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{secondwk}},
\code{\link{seecowrk}},
\code{\link{sethours}},
\code{\link{setthngs}},
\code{\link{shout}},
\code{\link{siteyrs}},
\code{\link{skipwork}},
\code{\link{slsmnshp}},
\code{\link{smallbig}},
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{splocnum}},
\code{\link{spvtrfair}},
\code{\link{spwrkslf}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{stress}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supervis}},
\code{\link{suphelp}},
\code{\link{supnum}},
\code{\link{supothrs}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{teamsafe}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timeoff}},
\code{\link{toldwork}},
\code{\link{toofewwk}},
\code{\link{totalnum}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{training}},
\code{\link{trainops}},
\code{\link{trainsch}},
\code{\link{trdestck}},
\code{\link{treatres}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{typeorg}},
\code{\link{unhlthy}},
\code{\link{unmanrel}},
\code{\link{unpleznt}},
\code{\link{useskill}},
\code{\link{usetech}},
\code{\link{wantjob}},
\code{\link{wantjob1}},
\code{\link{weekly}},
\code{\link{wherewrk}},
\code{\link{whybiz}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkbhvrs}},
\code{\link{wkbonus}},
\code{\link{wkcomptr}},
\code{\link{wkcontct}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkpraise}},
\code{\link{wksmooth}},
\code{\link{wksub}},
\code{\link{wksup}},
\code{\link{wktenure}},
\code{\link{wktopsat}},
\code{\link{work3yrs}},
\code{\link{workdiff}},
\code{\link{workdy}},
\code{\link{workfast}},
\code{\link{workless}},
\code{\link{worryjob}},
\code{\link{wrkindp}},
\code{\link{wrkoth}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrktime}},
\code{\link{xhaustn}},
\code{\link{yearbiz}},
\code{\link{yearly}},
\code{\link{yearsjob}},
\code{\link{yearval}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Computers And Internet}
\concept{Global Economics}
\concept{Job}
\keyword{variable}
