% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{wojob}
\alias{wojob}
\title{Since 1st job ever been without a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wojob
}
\details{
Question Since your first full-time job, has there ever been a time when you've been without a job and not looking for a job?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2157/vshow}{GSS Data Explorer page for \code{wojob}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab Total \cr
   1998 \tab 1455 \tab 495 \tab 854 \tab 28 \tab - \tab 2832 \cr
   Total \tab 1455 \tab 495 \tab 854 \tab 28 \tab 0 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{daily}},
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{locyrs}},
\code{\link{monthly}},
\code{\link{numemp}},
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{trainsch}},
\code{\link{trusting}},
\code{\link{typeorg}},
\code{\link{weekly}},
\code{\link{wojobyrs}},
\code{\link{wrkoth}},
\code{\link{yearly}}

Other Unemployment: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{econsup}},
\code{\link{evjob}},
\code{\link{findwork}},
\code{\link{getjob}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{lookjob}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unemp}},
\code{\link{unemp5}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrel1}},
\code{\link{unrel4}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{wantads}},
\code{\link{wantjob2}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyjbend}},
\code{\link{whyleave}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wojobyrs}},
\code{\link{work10}},
\code{\link{work7}},
\code{\link{work9}},
\code{\link{yrjobend}}
}
\concept{Jobs}
\concept{Unemployment}
\keyword{variable}
