% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_578.R
\name{worda}
\alias{worda}
\title{Word a}
\source{
General Social Survey https://gss.norc.org
}
\description{
worda
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.
A. Word A
}
\section{Values}{

\itemize{
\item \verb{[0]} incorrect answer
\item \verb{[1]} correct answer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024 \cr
   A/B/C \tab 1998, 2000, 2004, 2006 \cr
   Full \tab 1974, 1976, 1978, 1982, 1984, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5295/vshow}{GSS Data Explorer page for \code{worda}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab incorrect answer \tab correct answer \tab no answer \tab not available in this year \tab Total \cr
   1974 \tab - \tab 293 \tab 1116 \tab 75 \tab - \tab 1484 \cr
   1976 \tab - \tab 194 \tab 1198 \tab 107 \tab - \tab 1499 \cr
   1978 \tab - \tab 274 \tab 1185 \tab 73 \tab - \tab 1532 \cr
   1982 \tab - \tab 311 \tab 1384 \tab 165 \tab - \tab 1860 \cr
   1984 \tab - \tab 244 \tab 1124 \tab 105 \tab - \tab 1473 \cr
   1987 \tab - \tab 306 \tab 1324 \tab 189 \tab - \tab 1819 \cr
   1988 \tab 493 \tab 175 \tab 711 \tab 102 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 172 \tab 768 \tab 66 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 162 \tab 670 \tab 96 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 116 \tab 827 \tab 81 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 172 \tab 809 \tab 99 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 243 \tab 1563 \tab 171 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 266 \tab 1562 \tab 132 \tab - \tab 2904 \cr
   1998 \tab 1445 \tab 158 \tab 1115 \tab 114 \tab - \tab 2832 \cr
   2000 \tab 1398 \tab 157 \tab 1125 \tab 137 \tab - \tab 2817 \cr
   2004 \tab 1340 \tab 260 \tab 1168 \tab 44 \tab - \tab 2812 \cr
   2006 \tab 3041 \tab 207 \tab 1162 \tab 100 \tab - \tab 4510 \cr
   2008 \tab 845 \tab 194 \tab 941 \tab 43 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 262 \tab 1109 \tab 59 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 239 \tab 1012 \tab 51 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 296 \tab 1336 \tab 43 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 345 \tab 1489 \tab 54 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 331 \tab 1202 \tab 26 \tab - \tab 2348 \cr
   2022 \tab 2377 \tab 262 \tab 857 \tab 48 \tab - \tab 3544 \cr
   2024 \tab 2235 \tab 254 \tab 788 \tab 32 \tab - \tab 3309 \cr
   Total \tab 21044 \tab 5893 \tab 27545 \tab 2212 \tab 0 \tab 56694 \cr
}
}

\seealso{
Other Vocabulary: 
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Knowledge: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{anheat}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{boyorgrl}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{electron}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{governor}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{h2olife}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{nanoben}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{poleff20}},
\code{\link{radioact}},
\code{\link{schlhead}},
\code{\link{scifrom}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{seeksci}},
\code{\link{seesand}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{upbreath}},
\code{\link{usrep}},
\code{\link{viruses}},
\code{\link{weighing}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Vocabulary Test: 
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}}
}
\concept{Knowledge}
\concept{Vocabulary}
\concept{Vocabulary Test}
\keyword{variable}
