% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_578.R
\name{wordd}
\alias{wordd}
\title{Word d}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordd
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.
D. Word D
}
\section{Values}{

\itemize{
\item \verb{[0]} incorrect answer
\item \verb{[1]} correct answer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024 \cr
   A/B/C \tab 1998, 2000, 2004, 2006 \cr
   Full \tab 1974, 1976, 1978, 1982, 1984, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5298/vshow}{GSS Data Explorer page for \code{wordd}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab incorrect answer \tab correct answer \tab no answer \tab not available in this year \tab Total \cr
   1974 \tab - \tab 73 \tab 1359 \tab 52 \tab - \tab 1484 \cr
   1976 \tab - \tab 88 \tab 1304 \tab 107 \tab - \tab 1499 \cr
   1978 \tab - \tab 111 \tab 1347 \tab 74 \tab - \tab 1532 \cr
   1982 \tab - \tab 123 \tab 1556 \tab 181 \tab - \tab 1860 \cr
   1984 \tab - \tab 73 \tab 1299 \tab 101 \tab - \tab 1473 \cr
   1987 \tab - \tab 115 \tab 1496 \tab 208 \tab - \tab 1819 \cr
   1988 \tab 493 \tab 56 \tab 834 \tab 98 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 46 \tab 894 \tab 66 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 37 \tab 793 \tab 98 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 42 \tab 904 \tab 78 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 57 \tab 941 \tab 82 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 71 \tab 1747 \tab 159 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 81 \tab 1753 \tab 126 \tab - \tab 2904 \cr
   1998 \tab 1445 \tab 40 \tab 1229 \tab 118 \tab - \tab 2832 \cr
   2000 \tab 1398 \tab 69 \tab 1211 \tab 139 \tab - \tab 2817 \cr
   2004 \tab 1340 \tab 79 \tab 1337 \tab 56 \tab - \tab 2812 \cr
   2006 \tab 3041 \tab 59 \tab 1310 \tab 100 \tab - \tab 4510 \cr
   2008 \tab 845 \tab 50 \tab 1087 \tab 41 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 64 \tab 1308 \tab 58 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 68 \tab 1194 \tab 40 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 77 \tab 1563 \tab 35 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 92 \tab 1758 \tab 38 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 74 \tab 1457 \tab 28 \tab - \tab 2348 \cr
   2022 \tab 1178 \tab 91 \tab 2190 \tab 85 \tab - \tab 3544 \cr
   2024 \tab 1126 \tab 81 \tab 2053 \tab 49 \tab - \tab 3309 \cr
   Total \tab 18736 \tab 1817 \tab 33924 \tab 2217 \tab 0 \tab 56694 \cr
}
}

\seealso{
Other Vocabulary: 
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Knowledge: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{anheat}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{boyorgrl}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{electron}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{governor}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{h2olife}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{nanoben}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{poleff20}},
\code{\link{radioact}},
\code{\link{schlhead}},
\code{\link{scifrom}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{seeksci}},
\code{\link{seesand}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{upbreath}},
\code{\link{usrep}},
\code{\link{viruses}},
\code{\link{weighing}},
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Vocabulary Test: 
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordj}},
\code{\link{wordsum}}
}
\concept{Knowledge}
\concept{Vocabulary}
\concept{Vocabulary Test}
\keyword{variable}
