% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{wordj}
\alias{wordj}
\title{Word j}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordj
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.
J. Word J
}
\section{Values}{

\itemize{
\item \verb{[0]} incorrect answer
\item \verb{[1]} correct answer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024 \cr
   A/B/C \tab 1998, 2000, 2004, 2006 \cr
   Full \tab 1974, 1976, 1978, 1982, 1984, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5304/vshow}{GSS Data Explorer page for \code{wordj}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab incorrect answer \tab correct answer \tab no answer \tab not available in this year \tab Total \cr
   1974 \tab - \tab 1039 \tab 340 \tab 105 \tab - \tab 1484 \cr
   1976 \tab - \tab 1022 \tab 329 \tab 148 \tab - \tab 1499 \cr
   1978 \tab - \tab 1073 \tab 336 \tab 123 \tab - \tab 1532 \cr
   1982 \tab - \tab 1257 \tab 343 \tab 260 \tab - \tab 1860 \cr
   1984 \tab - \tab 1018 \tab 297 \tab 158 \tab - \tab 1473 \cr
   1987 \tab - \tab 1188 \tab 343 \tab 288 \tab - \tab 1819 \cr
   1988 \tab 493 \tab 669 \tab 180 \tab 139 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 659 \tab 225 \tab 122 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 534 \tab 226 \tab 168 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 667 \tab 214 \tab 143 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 692 \tab 250 \tab 138 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 1276 \tab 431 \tab 270 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 1225 \tab 456 \tab 279 \tab - \tab 2904 \cr
   1998 \tab 1445 \tab 793 \tab 355 \tab 239 \tab - \tab 2832 \cr
   2000 \tab 1398 \tab 858 \tab 297 \tab 264 \tab - \tab 2817 \cr
   2004 \tab 1340 \tab 993 \tab 397 \tab 82 \tab - \tab 2812 \cr
   2006 \tab 3041 \tab 937 \tab 357 \tab 175 \tab - \tab 4510 \cr
   2008 \tab 845 \tab 812 \tab 274 \tab 92 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 979 \tab 345 \tab 106 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 895 \tab 311 \tab 96 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 1183 \tab 414 \tab 78 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 1320 \tab 496 \tab 72 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 1099 \tab 409 \tab 51 \tab - \tab 2348 \cr
   2022 \tab 1179 \tab 1384 \tab 838 \tab 143 \tab - \tab 3544 \cr
   2024 \tab 1126 \tab 1245 \tab 837 \tab 101 \tab - \tab 3309 \cr
   Total \tab 18737 \tab 24817 \tab 9300 \tab 3840 \tab 0 \tab 56694 \cr
}
}

\seealso{
Other Vocabulary: 
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Knowledge: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{anheat}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{boyorgrl}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{electron}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{governor}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{h2olife}},
\code{\link{hotcore}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{nanoben}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{nosun}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{poleff20}},
\code{\link{radioact}},
\code{\link{schlhead}},
\code{\link{scifrom}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scistudy}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{scitext}},
\code{\link{seeksci}},
\code{\link{seesand}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen1}},
\code{\link{upbreath}},
\code{\link{usrep}},
\code{\link{viruses}},
\code{\link{weighing}},
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordsum}},
\code{\link{wordsum7}},
\code{\link{wordsumnew}},
\code{\link{wordsumold}}

Other Vocabulary Test: 
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}},
\code{\link{wordh}},
\code{\link{wordi}},
\code{\link{wordsum}}
}
\concept{Knowledge}
\concept{Vocabulary}
\concept{Vocabulary Test}
\keyword{variable}
