% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{wordk}
\alias{wordk}
\title{Word k}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordk
}
\details{
Question Please indicate the word that comes closest to the meaning of the word in capital letters.
}
\section{Values}{

\itemize{
\item \verb{[0]} incorrect answer
\item \verb{[1]} correct answer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7761/vshow}{GSS Data Explorer page for \code{wordk}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab incorrect answer \tab correct answer \tab no answer \tab Total \cr
   2022 \tab 2345 \tab 292 \tab 817 \tab 90 \tab 3544 \cr
   Total \tab 2345 \tab 292 \tab 817 \tab 90 \tab 3544 \cr
}
}

\keyword{variable}
