% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{wordsum}
\alias{wordsum}
\title{Number words correct in vocabulary test}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordsum
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.
K. Total number of correct words.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/- \tab partial \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5305/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab not available in this year \tab no answer \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 10 \tab 21 \tab 112 \tab 53 \tab 96 \tab 160 \tab 260 \tab 282 \tab 196 \tab 147 \tab 112 \tab 35 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 6 \tab 28 \tab 109 \tab 38 \tab 114 \tab 159 \tab 229 \tab 283 \tab 209 \tab 133 \tab 127 \tab 64 \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 10 \tab 37 \tab 107 \tab 48 \tab 109 \tab 164 \tab 226 \tab 307 \tab 229 \tab 141 \tab 108 \tab 46 \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 17 \tab 37 \tab 86 \tab 87 \tab 135 \tab 192 \tab 301 \tab 342 \tab 260 \tab 140 \tab 131 \tab 132 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 11 \tab 25 \tab 90 \tab 53 \tab 86 \tab 158 \tab 227 \tab 259 \tab 217 \tab 153 \tab 117 \tab 77 \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 24 \tab 38 \tab 85 \tab 54 \tab 144 \tab 209 \tab 257 \tab 370 \tab 219 \tab 151 \tab 118 \tab 150 \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 6 \tab 21 \tab 46 \tab 39 \tab 59 \tab 90 \tab 166 \tab 211 \tab 137 \tab 79 \tab 58 \tab 76 \tab - \tab - \tab 1481 \cr
   1989 \tab 531 \tab 14 \tab 21 \tab 51 \tab 43 \tab 52 \tab 95 \tab 145 \tab 219 \tab 147 \tab 113 \tab 71 \tab 35 \tab - \tab - \tab 1537 \cr
   1990 \tab 444 \tab 5 \tab 16 \tab 63 \tab 27 \tab 55 \tab 86 \tab 124 \tab 168 \tab 143 \tab 96 \tab 70 \tab 75 \tab - \tab - \tab 1372 \cr
   1991 \tab 493 \tab 1 \tab 18 \tab 55 \tab 38 \tab 44 \tab 98 \tab 158 \tab 206 \tab 170 \tab 88 \tab 85 \tab 63 \tab - \tab - \tab 1517 \cr
   1993 \tab 526 \tab 4 \tab 33 \tab 50 \tab 38 \tab 59 \tab 114 \tab 143 \tab 227 \tab 157 \tab 113 \tab 90 \tab 52 \tab - \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 7 \tab 35 \tab 113 \tab 43 \tab 96 \tab 181 \tab 309 \tab 415 \tab 304 \tab 196 \tab 152 \tab 126 \tab - \tab - \tab 2992 \cr
   1996 \tab 944 \tab 10 \tab 29 \tab 95 \tab 68 \tab 103 \tab 186 \tab 310 \tab 416 \tab 296 \tab 212 \tab 146 \tab 89 \tab - \tab - \tab 2904 \cr
   1998 \tab 1445 \tab 8 \tab 18 \tab 78 \tab 32 \tab 81 \tab 113 \tab 212 \tab 316 \tab 201 \tab 136 \tab 111 \tab 81 \tab - \tab - \tab 2832 \cr
   2000 \tab 1398 \tab 9 \tab 27 \tab 72 \tab 41 \tab 74 \tab 132 \tab 220 \tab 289 \tab 206 \tab 147 \tab 97 \tab 105 \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1340 \tab 20 \tab 19 \tab 74 \tab 27 \tab 54 \tab 137 \tab 219 \tab 324 \tab 270 \tab 170 \tab 125 \tab 33 \tab - \tab - \tab 2812 \cr
   2006 \tab 3041 \tab 10 \tab 22 \tab 48 \tab 38 \tab 59 \tab 121 \tab 224 \tab 315 \tab 241 \tab 195 \tab 118 \tab 78 \tab - \tab - \tab 4510 \cr
   2008 \tab 845 \tab 11 \tab 20 \tab 44 \tab 36 \tab 56 \tab 115 \tab 222 \tab 266 \tab 196 \tab 113 \tab 81 \tab 18 \tab - \tab - \tab 2023 \cr
   2010 \tab 614 \tab 4 \tab 22 \tab 67 \tab 53 \tab 79 \tab 138 \tab 213 \tab 337 \tab 211 \tab 151 \tab 112 \tab 43 \tab - \tab - \tab 2044 \cr
   2012 \tab 672 \tab 10 \tab 20 \tab 55 \tab 45 \tab 69 \tab 142 \tab 215 \tab 306 \tab 193 \tab 137 \tab 86 \tab 24 \tab - \tab - \tab 1974 \cr
   2014 \tab 863 \tab 5 \tab 34 \tab 53 \tab 46 \tab 90 \tab 153 \tab 277 \tab 400 \tab 284 \tab 195 \tab 118 \tab 20 \tab - \tab - \tab 2538 \cr
   2016 \tab 979 \tab 12 \tab 20 \tab 63 \tab 44 \tab 100 \tab 174 \tab 356 \tab 414 \tab 339 \tab 226 \tab 115 \tab 25 \tab - \tab - \tab 2867 \cr
   2018 \tab 789 \tab 14 \tab 26 \tab 52 \tab 45 \tab 100 \tab 151 \tab 272 \tab 338 \tab 266 \tab 191 \tab 92 \tab 12 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab 1168 \tab 13 \tab 48 \tab 146 \tab 74 \tab 111 \tab 193 \tab 305 \tab 425 \tab 358 \tab 347 \tab 281 \tab - \tab - \tab 75 \tab 3544 \cr
   2024 \tab 1126 \tab 15 \tab 50 \tab 113 \tab 55 \tab 133 \tab 178 \tab 278 \tab 380 \tab 330 \tab 333 \tab 293 \tab - \tab - \tab 25 \tab 3309 \cr
   Total \tab 33699 \tab 256 \tab 685 \tab 1927 \tab 1165 \tab 2158 \tab 3639 \tab 5868 \tab 7815 \tab 5779 \tab 4103 \tab 3014 \tab 1459 \tab 4032 \tab 100 \tab 75699 \cr
}
}

\seealso{
Other Vocabulary: 
\code{\link{wordi}},
\code{\link{wordj}}

Other Knowledge: 
\code{\link{wordi}},
\code{\link{wordj}}

Other Vocabulary Test: 
\code{\link{wordi}},
\code{\link{wordj}}
}
\concept{Knowledge}
\concept{Vocabulary}
\concept{Vocabulary Test}
\keyword{variable}
