% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_690.R
\name{wordsum7}
\alias{wordsum7}
\title{Number words correct in vocabulary test (7 common items)}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordsum7
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.

K. Total number of correct words.
NOTE: This variable only covers the 7 common vocabulary items from the original and experimental versions of WORDSUM. For more details, see the Codebook.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/-/- \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1974, 1976, 1978, 1982, 1984, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8260/vshow}{GSS Data Explorer page for \code{wordsum7}}.

Counts by year:\tabular{llllllllllll}{
   year \tab no answer \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab iap \tab Total \cr
   1974 \tab 35 \tab 20 \tab 54 \tab 159 \tab 277 \tab 386 \tab 249 \tab 152 \tab 152 \tab - \tab 1484 \cr
   1976 \tab 64 \tab 23 \tab 77 \tab 165 \tab 241 \tab 388 \tab 250 \tab 153 \tab 138 \tab - \tab 1499 \cr
   1978 \tab 46 \tab 29 \tab 75 \tab 180 \tab 246 \tab 409 \tab 256 \tab 155 \tab 136 \tab - \tab 1532 \cr
   1982 \tab 132 \tab 39 \tab 93 \tab 214 \tab 298 \tab 504 \tab 273 \tab 188 \tab 119 \tab - \tab 1860 \cr
   1984 \tab 77 \tab 23 \tab 63 \tab 149 \tab 243 \tab 375 \tab 261 \tab 157 \tab 125 \tab - \tab 1473 \cr
   1987 \tab 150 \tab 44 \tab 100 \tab 185 \tab 300 \tab 483 \tab 263 \tab 176 \tab 118 \tab - \tab 1819 \cr
   1988 \tab 569 \tab 19 \tab 42 \tab 95 \tab 157 \tab 284 \tab 153 \tab 97 \tab 65 \tab - \tab 1481 \cr
   1989 \tab 553 \tab 35 \tab 51 \tab 87 \tab 149 \tab 298 \tab 182 \tab 122 \tab 60 \tab - \tab 1537 \cr
   1990 \tab 519 \tab 12 \tab 39 \tab 70 \tab 137 \tab 234 \tab 177 \tab 102 \tab 82 \tab - \tab 1372 \cr
   1991 \tab 556 \tab 14 \tab 38 \tab 95 \tab 153 \tab 302 \tab 171 \tab 111 \tab 77 \tab - \tab 1517 \cr
   1993 \tab 576 \tab 20 \tab 56 \tab 77 \tab 178 \tab 316 \tab 178 \tab 137 \tab 68 \tab - \tab 1606 \cr
   1994 \tab 888 \tab 29 \tab 68 \tab 145 \tab 295 \tab 614 \tab 337 \tab 221 \tab 151 \tab 244 \tab 2992 \cr
   1996 \tab 564 \tab 26 \tab 69 \tab 159 \tab 330 \tab 571 \tab 386 \tab 201 \tab 129 \tab 469 \tab 2904 \cr
   1998 \tab 1028 \tab 15 \tab 55 \tab 102 \tab 199 \tab 423 \tab 246 \tab 162 \tab 104 \tab 498 \tab 2832 \cr
   2000 \tab 1049 \tab 22 \tab 67 \tab 112 \tab 209 \tab 396 \tab 273 \tab 139 \tab 96 \tab 454 \tab 2817 \cr
   2002 \tab 2301 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 464 \tab 2765 \cr
   2004 \tab 939 \tab 33 \tab 37 \tab 91 \tab 190 \tab 456 \tab 335 \tab 189 \tab 108 \tab 434 \tab 2812 \cr
   2006 \tab 2608 \tab 24 \tab 38 \tab 82 \tab 218 \tab 448 \tab 312 \tab 195 \tab 74 \tab 511 \tab 4510 \cr
   2008 \tab 169 \tab 25 \tab 40 \tab 92 \tab 196 \tab 381 \tab 232 \tab 128 \tab 66 \tab 694 \tab 2023 \cr
   2010 \tab 43 \tab 16 \tab 57 \tab 107 \tab 228 \tab 447 \tab 247 \tab 197 \tab 88 \tab 614 \tab 2044 \cr
   2012 \tab 22 \tab 30 \tab 42 \tab 116 \tab 203 \tab 409 \tab 253 \tab 149 \tab 78 \tab 672 \tab 1974 \cr
   2014 \tab 18 \tab 24 \tab 64 \tab 123 \tab 271 \tab 508 \tab 390 \tab 196 \tab 81 \tab 863 \tab 2538 \cr
   2016 \tab 23 \tab 25 \tab 59 \tab 125 \tab 313 \tab 600 \tab 425 \tab 227 \tab 91 \tab 979 \tab 2867 \cr
   2018 \tab 12 \tab 32 \tab 57 \tab 121 \tab 261 \tab 465 \tab 327 \tab 203 \tab 81 \tab 789 \tab 2348 \cr
   2021 \tab 2717 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1315 \tab 4032 \cr
   2022 \tab 75 \tab 27 \tab 72 \tab 142 \tab 292 \tab 628 \tab 526 \tab 419 \tab 195 \tab 1168 \tab 3544 \cr
   2024 \tab 1151 \tab 23 \tab 73 \tab 143 \tab 267 \tab 568 \tab 494 \tab 433 \tab 157 \tab - \tab 3309 \cr
   Total \tab 16884 \tab 629 \tab 1486 \tab 3136 \tab 5851 \tab 10893 \tab 7196 \tab 4609 \tab 2639 \tab 10168 \tab 63491 \cr
}
}

\seealso{
Other Vocabulary: 
\code{\link{wordsumold}}

Other Knowledge: 
\code{\link{wordsumold}}
}
\concept{Knowledge}
\concept{Vocabulary}
\keyword{variable}
