% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_691.R
\name{wordsumnew}
\alias{wordsumnew}
\title{Number words correct in vocabulary test (experiment items)}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordsumnew
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.

K. Total number of correct words.
NOTE: This variable contains the 10 experimental vocabulary items from WORDSUM. For more details, see the Codebook. For more details, see the Codebook.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2022 \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8261/vshow}{GSS Data Explorer page for \code{wordsumnew}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab iap \tab no answer \tab Total \cr
   2022 \tab - \tab 6 \tab 27 \tab 94 \tab 34 \tab 65 \tab 102 \tab 152 \tab 168 \tab 156 \tab 191 \tab 170 \tab 2339 \tab 40 \tab 3544 \cr
   2024 \tab - \tab 6 \tab 31 \tab 66 \tab 30 \tab 74 \tab 82 \tab 136 \tab 162 \tab 158 \tab 176 \tab 175 \tab 1626 \tab 587 \tab 3309 \cr
   Total \tab 0 \tab 12 \tab 58 \tab 160 \tab 64 \tab 139 \tab 184 \tab 288 \tab 330 \tab 314 \tab 367 \tab 345 \tab 3965 \tab 627 \tab 6853 \cr
}
}

\concept{Knowledge}
\concept{Vocabulary}
\keyword{variable}
