% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_690.R
\name{wordsumold}
\alias{wordsumold}
\title{Number words correct in vocabulary test (original items)}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordsumold
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.

K. Total number of correct words.
NOTE: This variable contains the 10 original vocabulary items from WORDSUM. For more details, see the Codebook. For more details, see the Codebook.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/-/- \tab 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1974, 1976, 1978, 1982, 1984, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8262/vshow}{GSS Data Explorer page for \code{wordsumold}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab no answer \tab 0 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab iap \tab Total \cr
   1974 \tab 35 \tab 10 \tab 21 \tab 112 \tab 53 \tab 96 \tab 160 \tab 260 \tab 282 \tab 196 \tab 147 \tab 112 \tab - \tab 1484 \cr
   1976 \tab 64 \tab 6 \tab 28 \tab 109 \tab 38 \tab 114 \tab 159 \tab 229 \tab 283 \tab 209 \tab 133 \tab 127 \tab - \tab 1499 \cr
   1978 \tab 46 \tab 10 \tab 37 \tab 107 \tab 48 \tab 109 \tab 164 \tab 226 \tab 307 \tab 229 \tab 141 \tab 108 \tab - \tab 1532 \cr
   1982 \tab 132 \tab 17 \tab 37 \tab 86 \tab 87 \tab 135 \tab 192 \tab 301 \tab 342 \tab 260 \tab 140 \tab 131 \tab - \tab 1860 \cr
   1984 \tab 77 \tab 11 \tab 25 \tab 90 \tab 53 \tab 86 \tab 158 \tab 227 \tab 259 \tab 217 \tab 153 \tab 117 \tab - \tab 1473 \cr
   1987 \tab 150 \tab 24 \tab 38 \tab 85 \tab 54 \tab 144 \tab 209 \tab 257 \tab 370 \tab 219 \tab 151 \tab 118 \tab - \tab 1819 \cr
   1988 \tab 569 \tab 6 \tab 21 \tab 46 \tab 39 \tab 59 \tab 90 \tab 166 \tab 211 \tab 137 \tab 79 \tab 58 \tab - \tab 1481 \cr
   1989 \tab 553 \tab 27 \tab 21 \tab 51 \tab 43 \tab 52 \tab 95 \tab 145 \tab 219 \tab 147 \tab 113 \tab 71 \tab - \tab 1537 \cr
   1990 \tab 519 \tab 5 \tab 16 \tab 63 \tab 27 \tab 55 \tab 86 \tab 124 \tab 168 \tab 143 \tab 96 \tab 70 \tab - \tab 1372 \cr
   1991 \tab 556 \tab 1 \tab 18 \tab 55 \tab 38 \tab 44 \tab 98 \tab 158 \tab 206 \tab 170 \tab 88 \tab 85 \tab - \tab 1517 \cr
   1993 \tab 576 \tab 6 \tab 33 \tab 50 \tab 38 \tab 59 \tab 114 \tab 143 \tab 227 \tab 157 \tab 113 \tab 90 \tab - \tab 1606 \cr
   1994 \tab 888 \tab 16 \tab 35 \tab 113 \tab 43 \tab 96 \tab 181 \tab 309 \tab 415 \tab 304 \tab 196 \tab 152 \tab 244 \tab 2992 \cr
   1996 \tab 564 \tab 10 \tab 29 \tab 95 \tab 68 \tab 103 \tab 186 \tab 310 \tab 416 \tab 296 \tab 212 \tab 146 \tab 469 \tab 2904 \cr
   1998 \tab 1028 \tab 8 \tab 18 \tab 78 \tab 32 \tab 81 \tab 113 \tab 212 \tab 316 \tab 201 \tab 136 \tab 111 \tab 498 \tab 2832 \cr
   2000 \tab 1049 \tab 9 \tab 27 \tab 72 \tab 41 \tab 74 \tab 132 \tab 220 \tab 289 \tab 206 \tab 147 \tab 97 \tab 454 \tab 2817 \cr
   2002 \tab 2301 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 464 \tab 2765 \cr
   2004 \tab 939 \tab 20 \tab 19 \tab 74 \tab 27 \tab 54 \tab 137 \tab 219 \tab 324 \tab 270 \tab 170 \tab 125 \tab 434 \tab 2812 \cr
   2006 \tab 2608 \tab 10 \tab 22 \tab 48 \tab 38 \tab 59 \tab 121 \tab 224 \tab 315 \tab 241 \tab 195 \tab 118 \tab 511 \tab 4510 \cr
   2008 \tab 169 \tab 11 \tab 20 \tab 44 \tab 36 \tab 56 \tab 115 \tab 222 \tab 266 \tab 196 \tab 113 \tab 81 \tab 694 \tab 2023 \cr
   2010 \tab 43 \tab 4 \tab 22 \tab 67 \tab 53 \tab 79 \tab 138 \tab 213 \tab 337 \tab 211 \tab 151 \tab 112 \tab 614 \tab 2044 \cr
   2012 \tab 22 \tab 12 \tab 20 \tab 55 \tab 45 \tab 69 \tab 142 \tab 215 \tab 306 \tab 193 \tab 137 \tab 86 \tab 672 \tab 1974 \cr
   2014 \tab 18 \tab 7 \tab 34 \tab 53 \tab 46 \tab 90 \tab 153 \tab 277 \tab 400 \tab 284 \tab 195 \tab 118 \tab 863 \tab 2538 \cr
   2016 \tab 23 \tab 14 \tab 20 \tab 63 \tab 44 \tab 100 \tab 174 \tab 356 \tab 414 \tab 339 \tab 226 \tab 115 \tab 979 \tab 2867 \cr
   2018 \tab 12 \tab 14 \tab 26 \tab 52 \tab 45 \tab 100 \tab 151 \tab 272 \tab 338 \tab 266 \tab 191 \tab 92 \tab 789 \tab 2348 \cr
   2021 \tab 2717 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1315 \tab 4032 \cr
   2022 \tab 35 \tab 7 \tab 21 \tab 52 \tab 40 \tab 46 \tab 91 \tab 153 \tab 257 \tab 202 \tab 156 \tab 111 \tab 2373 \tab 3544 \cr
   2024 \tab 564 \tab 9 \tab 19 \tab 47 \tab 25 \tab 59 \tab 96 \tab 142 \tab 218 \tab 172 \tab 157 \tab 118 \tab 1683 \tab 3309 \cr
   Total \tab 16257 \tab 274 \tab 627 \tab 1767 \tab 1101 \tab 2019 \tab 3455 \tab 5580 \tab 7485 \tab 5465 \tab 3736 \tab 2669 \tab 13056 \tab 63491 \cr
}
}

\seealso{
Other Vocabulary: 
\code{\link{wordsum7}}

Other Knowledge: 
\code{\link{wordsum7}}
}
\concept{Knowledge}
\concept{Vocabulary}
\keyword{variable}
