% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{worecsup}
\alias{worecsup}
\title{Do you worry about losing economic support}
\source{
General Social Survey https://gss.norc.org
}
\description{
worecsup
}
\details{
Question To what extent, if at all, do you worry about the possibility of losing this main economic support?             
(Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} i worry a great deal
\item \code{2} i worry to some extent
\item \code{3} i worry a little
\item \code{4} i don't worry at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6350/vshow}{GSS Data Explorer page for \code{worecsup}}.
}

\seealso{
Other Work Orientation: 
\code{\link{econsup1}},
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrespct}},
\code{\link{unslfcon}}
}
\concept{Work Orientation}
\keyword{variable}
