% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_562.R
\name{work2}
\alias{work2}
\title{Being demoted or move to worse position}
\source{
General Social Survey https://gss.norc.org
}
\description{
work2
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
B. Second, what about work-related matters?  Did any of the following happen to you since (CURRENT MONTH), (1990/2003)?
2. Demoted or switched to a less favorable position.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5151/vshow}{GSS Data Explorer page for \code{work2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1991 \tab 500 \tab 26 \tab 946 \tab 45 \tab - \tab 1517 \cr
   2004 \tab 1472 \tab 32 \tab 1299 \tab 8 \tab 1 \tab 2812 \cr
   Total \tab 1972 \tab 58 \tab 2245 \tab 53 \tab 1 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth12}},
\code{\link{hlth13}},
\code{\link{hlth14}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{work1}},
\code{\link{work3}}
}
\concept{Negative Life Events}
\keyword{variable}
