% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{work30}
\alias{work30}
\title{R use work site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
work30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
D. Sites related to your work?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} 3-5 times
\item \code{4} more than 5 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2397/vshow}{GSS Data Explorer page for \code{work30}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab 1-2 times \tab 3-5 times \tab more than 5 times \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 262 \tab 108 \tab 82 \tab 215 \tab 75 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 426 \tab 137 \tab 89 \tab 338 \tab 93 \tab 1 \tab 2765 \cr
   2004 \tab 2122 \tab 303 \tab 96 \tab 79 \tab 200 \tab 12 \tab - \tab 2812 \cr
   Total \tab 5878 \tab 991 \tab 341 \tab 250 \tab 753 \tab 180 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{bookmark}},
\code{\link{catdrcty}},
\code{\link{educ30}},
\code{\link{emaillnk}},
\code{\link{fin30}},
\code{\link{govt30}},
\code{\link{hyperlnk}},
\code{\link{news30}},
\code{\link{schl30}}
}
\concept{Information Society}
\keyword{variable}
