% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_563.R
\name{work4}
\alias{work4}
\title{Being passed over for promotion}
\source{
General Social Survey https://gss.norc.org
}
\description{
work4
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
B. Second, what about work-related matters?  Did any of the following happen to you since (CURRENT MONTH), (1990/2003)?
4. Passed over for promotion.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5153/vshow}{GSS Data Explorer page for \code{work4}}.
}

\seealso{
Other Negative Life Events: 
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{work10}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}

Other Trauma: 
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{work10}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}
}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
