% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_563.R
\name{work5}
\alias{work5}
\title{Having trouble with one's boss}
\source{
General Social Survey https://gss.norc.org
}
\description{
work5
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
B. Second, what about work-related matters?  Did any of the following happen to you since (CURRENT MONTH), (1990/2003)?
5. Serious trouble with your boss.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5154/vshow}{GSS Data Explorer page for \code{work5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1991 \tab 500 \tab 42 \tab 929 \tab 46 \tab - \tab 1517 \cr
   2004 \tab 1472 \tab 60 \tab 1270 \tab 9 \tab 1 \tab 2812 \cr
   Total \tab 1972 \tab 102 \tab 2199 \tab 55 \tab 1 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{work10}},
\code{\link{work4}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}
}
\concept{Negative Life Events}
\keyword{variable}
