% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_563.R
\name{work6}
\alias{work6}
\title{Own business losing money or failing}
\source{
General Social Survey https://gss.norc.org
}
\description{
work6
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
B. Second, what about work-related matters?  Did any of the following happen to you since (CURRENT MONTH), (1990/2003)?
6. Your own business (farm or professional office) lost money or failed.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5155/vshow}{GSS Data Explorer page for \code{work6}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1991 \tab 500 \tab 20 \tab 951 \tab 46 \tab - \tab 1517 \cr
   2004 \tab 1472 \tab 45 \tab 1281 \tab 10 \tab 4 \tab 2812 \cr
   Total \tab 1972 \tab 65 \tab 2232 \tab 56 \tab 4 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{work10}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work7}},
\code{\link{work8}},
\code{\link{work9}}
}
\concept{Negative Life Events}
\keyword{variable}
