% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_563.R
\name{work9}
\alias{work9}
\title{Ones' spouse being unemployed}
\source{
General Social Survey https://gss.norc.org
}
\description{
work9
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
B. Second, what about work-related matters?  Did any of the following happen to you since (CURRENT MONTH), (1990/2003)?
9. Your (husband/wife)/partner was unemployed and looking for work for a month or more.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5158/vshow}{GSS Data Explorer page for \code{work9}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1991 \tab 500 \tab 55 \tab 911 \tab 2 \tab 49 \tab 1517 \cr
   2004 \tab 1472 \tab 104 \tab 1222 \tab 2 \tab 12 \tab 2812 \cr
   Total \tab 1972 \tab 159 \tab 2133 \tab 4 \tab 61 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{work10}},
\code{\link{work4}},
\code{\link{work5}},
\code{\link{work6}},
\code{\link{work7}},
\code{\link{work8}}
}
\concept{Negative Life Events}
\keyword{variable}
