% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{workasns}
\alias{workasns}
\title{Hard working - lazy}
\source{
General Social Survey https://gss.norc.org
}
\description{
workasns
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
B. The second set of characteristics asks if people in the group tend to be hard-working or if they tend to be lazy.
4. Asian Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - hard-working
\item \verb{[7]} 7 - lazy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1187/vshow}{GSS Data Explorer page for \code{workasns}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - hard-working \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - lazy \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1990 \tab - \tab 143 \tab 192 \tab 302 \tab 375 \tab 132 \tab 52 \tab 24 \tab 126 \tab 26 \tab - \tab 1372 \cr
   1994 \tab 1518 \tab 176 \tab 293 \tab 332 \tab 353 \tab 108 \tab 52 \tab 15 \tab 128 \tab 17 \tab - \tab 2992 \cr
   2000 \tab 1419 \tab 152 \tab 238 \tab 312 \tab 392 \tab 95 \tab 31 \tab 9 \tab 153 \tab 16 \tab - \tab 2817 \cr
   Total \tab 2937 \tab 471 \tab 723 \tab 946 \tab 1120 \tab 335 \tab 135 \tab 48 \tab 407 \tab 59 \tab 0 \tab 7181 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkcol}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}},
\code{\link{fareasns}},
\code{\link{fareblks}},
\code{\link{farehsps}},
\code{\link{farejews}},
\code{\link{fareso}},
\code{\link{farewhts}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influblk}},
\code{\link{influhsp}},
\code{\link{influjew}},
\code{\link{influso}},
\code{\link{influwht}},
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intlhsps}},
\code{\link{intljews}},
\code{\link{intlso}},
\code{\link{intlwhts}},
\code{\link{lessfare}},
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrjews}},
\code{\link{patrso}},
\code{\link{patrwhts}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}},
\code{\link{racwork}},
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{talkback}},
\code{\link{twoclass}},
\code{\link{violasns}},
\code{\link{violblks}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthjews}},
\code{\link{wlthso}},
\code{\link{wlthwhts}},
\code{\link{workblks}},
\code{\link{workfare}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Asian Americans: 
\code{\link{acqasian}},
\code{\link{acqfmasn}},
\code{\link{acqnhasn}},
\code{\link{acqvaasn}},
\code{\link{acqwkasn}},
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asngovt}},
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{comasn}},
\code{\link{fairasns}},
\code{\link{famasns}},
\code{\link{fareasns}},
\code{\link{feelasns}},
\code{\link{hspasn10}},
\code{\link{influasn}},
\code{\link{intlasns}},
\code{\link{knwasn}},
\code{\link{liveasns}},
\code{\link{marasian}},
\code{\link{patrasns}},
\code{\link{trtasian}},
\code{\link{usasn}},
\code{\link{violasns}},
\code{\link{wlthasns}}
}
\concept{Asian Americans}
\concept{Intergroup Relations}
\keyword{variable}
