% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{workblks}
\alias{workblks}
\title{Hard working - lazy}
\source{
General Social Survey https://gss.norc.org
}
\description{
workblks
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
B. The second set of characteristics asks if people in the group tend to be hard-working or if they tend to be lazy.
3. Blacks?
}
\section{Values}{

\itemize{
\item \code{1} 1 - hard-working
\item \code{7} 7 - lazy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1996, 1998, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1990, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1186/vshow}{GSS Data Explorer page for \code{workblks}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - hard-working \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - lazy \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 45 \tab 55 \tab 166 \tab 460 \tab 308 \tab 182 \tab 79 \tab 51 \tab 26 \tab - \tab 1372 \cr
   1994 \tab 1518 \tab 49 \tab 96 \tab 174 \tab 492 \tab 317 \tab 192 \tab 88 \tab 48 \tab 18 \tab - \tab 2992 \cr
   1996 \tab 1935 \tab 21 \tab 49 \tab 96 \tab 424 \tab 209 \tab 95 \tab 25 \tab 43 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 34 \tab 89 \tab 243 \tab 838 \tab 360 \tab 163 \tab 54 \tab 79 \tab 11 \tab - \tab 2832 \cr
   2000 \tab 486 \tab 69 \tab 119 \tab 300 \tab 967 \tab 459 \tab 222 \tab 79 \tab 91 \tab 25 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 37 \tab 44 \tab 135 \tab 373 \tab 189 \tab 66 \tab 34 \tab 23 \tab 7 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 31 \tab 43 \tab 102 \tab 452 \tab 161 \tab 69 \tab 24 \tab 19 \tab 5 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 57 \tab 74 \tab 230 \tab 917 \tab 404 \tab 174 \tab 73 \tab 50 \tab 13 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 27 \tab 55 \tab 151 \tab 608 \tab 275 \tab 112 \tab 44 \tab 38 \tab 19 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 53 \tab 61 \tab 173 \tab 667 \tab 269 \tab 118 \tab 40 \tab 34 \tab 15 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 26 \tab 46 \tab 151 \tab 629 \tab 254 \tab 107 \tab 46 \tab 23 \tab 20 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 42 \tab 69 \tab 210 \tab 853 \tab 296 \tab 118 \tab 53 \tab 22 \tab 12 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 66 \tab 100 \tab 243 \tab 939 \tab 332 \tab 118 \tab 55 \tab 21 \tab 14 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 45 \tab 88 \tab 219 \tab 786 \tab 236 \tab 106 \tab 32 \tab 26 \tab 21 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 202 \tab 209 \tab 497 \tab 1341 \tab 303 \tab 82 \tab 43 \tab 9 \tab 2 \tab 29 \tab 4032 \cr
   2022 \tab 1168 \tab 178 \tab 134 \tab 392 \tab 1147 \tab 307 \tab 101 \tab 59 \tab 31 \tab 17 \tab 10 \tab 3544 \cr
   Total \tab 18275 \tab 982 \tab 1331 \tab 3482 \tab 11893 \tab 4679 \tab 2025 \tab 828 \tab 608 \tab 232 \tab 39 \tab 44374 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthso}},
\code{\link{workasns}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
