% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{workdiff}
\alias{workdiff}
\title{R does numerous things on job}
\source{
General Social Survey https://gss.norc.org
}
\description{
workdiff
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
C. I get to do a number of different things on my job
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2776/vshow}{GSS Data Explorer page for \code{workdiff}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 750 \tab 836 \tab 158 \tab 32 \tab 7 \tab 13 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 726 \tab 780 \tab 188 \tab 21 \tab 7 \tab 12 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 478 \tab 563 \tab 101 \tab 20 \tab 2 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 434 \tab 679 \tab 113 \tab 16 \tab 3 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 497 \tab 761 \tab 134 \tab 22 \tab 2 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 713 \tab 1024 \tab 192 \tab 35 \tab 6 \tab 12 \tab 16 \tab 3544 \cr
   Total \tab 8336 \tab 3598 \tab 4643 \tab 886 \tab 146 \tab 27 \tab 97 \tab 16 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{hrsrelax}},
\code{\link{learnnew}},
\code{\link{secondwk}},
\code{\link{whywkhme}},
\code{\link{wkvsfam}},
\code{\link{workfast}},
\code{\link{wrkhome}}
}
\concept{Quality of Working Life}
\keyword{variable}
