% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_320.R
\name{workfor1}
\alias{workfor1}
\title{R work for whom}
\source{
General Social Survey https://gss.norc.org
}
\description{
workfor1
}
\details{
Question Do you work for a for-profit company, a non-profit or not-for-profit organization, or for the government or a government agency?
}
\section{Values}{

\itemize{
\item \code{1} for-profit company
\item \code{2} non-profit or not-for-profit organization
\item \code{3} government or government agency
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2851/vshow}{GSS Data Explorer page for \code{workfor1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab for-profit company \tab non-profit or not-for-profit organization \tab government or government agency \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2014 \tab 1765 \tab 514 \tab 115 \tab 126 \tab 13 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 936 \tab 201 \tab 229 \tab 47 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 2569 \tab 594 \tab 143 \tab 145 \tab 81 \tab 8 \tab 4 \tab 3544 \cr
   Total \tab 5234 \tab 2044 \tab 459 \tab 500 \tab 141 \tab 48 \tab 4 \tab 8430 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{ownstock}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{workfor}}
}
\concept{Employee Compensation}
\keyword{variable}
