% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{workhard}
\alias{workhard}
\title{To work hard}
\source{
General Social Survey https://gss.norc.org
}
\description{
workhard
}
\details{
Question If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life?  (INTERVIEWER: READ CHOICES)
A. Which comes next in importance?
B. Which comes third?
C. Which comes fourth?
4. To work hard
}
\section{Values}{

\itemize{
\item \code{1} first
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/516/vshow}{GSS Data Explorer page for \code{workhard}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab first \tab second \tab third \tab fourth \tab fifth \tab no answer \tab Total \cr
   1986 \tab 730 \tab 81 \tab 244 \tab 238 \tab 129 \tab 40 \tab 8 \tab 1470 \cr
   1987 \tab - \tab 214 \tab 621 \tab 596 \tab 304 \tab 65 \tab 19 \tab 1819 \cr
   1988 \tab 484 \tab 140 \tab 303 \tab 324 \tab 174 \tab 36 \tab 20 \tab 1481 \cr
   1989 \tab 502 \tab 138 \tab 335 \tab 347 \tab 158 \tab 22 \tab 35 \tab 1537 \cr
   1990 \tab 473 \tab 139 \tab 301 \tab 294 \tab 116 \tab 21 \tab 28 \tab 1372 \cr
   1991 \tab 500 \tab 147 \tab 355 \tab 321 \tab 144 \tab 15 \tab 35 \tab 1517 \cr
   1993 \tab 549 \tab 143 \tab 368 \tab 369 \tab 129 \tab 22 \tab 26 \tab 1606 \cr
   1994 \tab 981 \tab 302 \tab 768 \tab 623 \tab 255 \tab 30 \tab 33 \tab 2992 \cr
   1996 \tab 979 \tab 327 \tab 697 \tab 611 \tab 229 \tab 31 \tab 30 \tab 2904 \cr
   1998 \tab 921 \tab 325 \tab 691 \tab 598 \tab 238 \tab 28 \tab 31 \tab 2832 \cr
   2000 \tab 921 \tab 315 \tab 662 \tab 596 \tab 256 \tab 34 \tab 33 \tab 2817 \cr
   2002 \tab 1853 \tab 190 \tab 314 \tab 288 \tab 90 \tab 17 \tab 13 \tab 2765 \cr
   2004 \tab 1936 \tab 180 \tab 287 \tab 296 \tab 95 \tab 10 \tab 8 \tab 2812 \cr
   2006 \tab 3508 \tab 186 \tab 366 \tab 304 \tab 124 \tab 15 \tab 7 \tab 4510 \cr
   2008 \tab 658 \tab 295 \tab 454 \tab 420 \tab 157 \tab 25 \tab 14 \tab 2023 \cr
   2010 \tab 667 \tab 338 \tab 467 \tab 400 \tab 121 \tab 23 \tab 28 \tab 2044 \cr
   2012 \tab 636 \tab 326 \tab 451 \tab 399 \tab 127 \tab 20 \tab 15 \tab 1974 \cr
   2014 \tab 851 \tab 425 \tab 609 \tab 485 \tab 137 \tab 14 \tab 17 \tab 2538 \cr
   2016 \tab 911 \tab 510 \tab 696 \tab 560 \tab 146 \tab 22 \tab 22 \tab 2867 \cr
   2018 \tab 785 \tab 415 \tab 515 \tab 483 \tab 114 \tab 15 \tab 21 \tab 2348 \cr
   2021 \tab 1360 \tab 624 \tab 934 \tab 894 \tab 116 \tab 5 \tab 99 \tab 4032 \cr
   2022 \tab 1173 \tab 554 \tab 827 \tab 724 \tab 166 \tab 23 \tab 77 \tab 3544 \cr
   Total \tab 21378 \tab 6314 \tab 11265 \tab 10170 \tab 3525 \tab 533 \tab 619 \tab 53804 \cr
}
}

\seealso{
Other Core: 
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{helpoth}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socrel}},
\code{\link{thnkself}},
\code{\link{youngen}}
}
\concept{Core}
\keyword{variable}
