% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{workhr}
\alias{workhr}
\title{Hours r usually works}
\source{
General Social Survey https://gss.norc.org
}
\description{
workhr
}
\details{
Question What hours do you usually work--days, evenings, or nights?  CIRCLE ONE CODE.
}
\section{Values}{

\itemize{
\item \code{1} reg day
\item \code{2} eve
\item \code{3} nights
\item \code{4} varies
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5288/vshow}{GSS Data Explorer page for \code{workhr}}.
}

\seealso{
Other Military Service & Work: 
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{vetkind}},
\code{\link{vetyears}},
\code{\link{workdy}}
}
\concept{Military Service & Work}
\keyword{variable}
