% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{workhsps}
\alias{workhsps}
\title{Hard working -lazy}
\source{
General Social Survey https://gss.norc.org
}
\description{
workhsps
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
B. The second set of characteristics asks if people in the group tend to be hard-working or if they tend to be lazy.
5. Hispanic Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - hard-working
\item \verb{[7]} 7 - lazy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1990, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1188/vshow}{GSS Data Explorer page for \code{workhsps}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - hard-working \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - lazy \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 53 \tab 73 \tab 210 \tab 449 \tab 269 \tab 148 \tab 37 \tab 106 \tab 27 \tab - \tab 1372 \cr
   1994 \tab 1518 \tab 53 \tab 99 \tab 212 \tab 506 \tab 296 \tab 121 \tab 54 \tab 115 \tab 18 \tab - \tab 2992 \cr
   2000 \tab 1419 \tab 93 \tab 148 \tab 236 \tab 508 \tab 181 \tab 71 \tab 25 \tab 120 \tab 16 \tab - \tab 2817 \cr
   2018 \tab 789 \tab 162 \tab 275 \tab 357 \tab 560 \tab 109 \tab 33 \tab 15 \tab 29 \tab 19 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 519 \tab 442 \tab 625 \tab 934 \tab 116 \tab 24 \tab 19 \tab 9 \tab 2 \tab 27 \tab 4032 \cr
   2022 \tab 1168 \tab 452 \tab 391 \tab 499 \tab 778 \tab 118 \tab 44 \tab 22 \tab 39 \tab 18 \tab 15 \tab 3544 \cr
   2024 \tab 1126 \tab 427 \tab 387 \tab 473 \tab 666 \tab 95 \tab 24 \tab 16 \tab 51 \tab 17 \tab 27 \tab 3309 \cr
   Total \tab 7335 \tab 1759 \tab 1815 \tab 2612 \tab 4401 \tab 1184 \tab 465 \tab 188 \tab 469 \tab 117 \tab 69 \tab 20414 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthso}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workjews}},
\code{\link{workso}},
\code{\link{workwhts}}

Other Hispanics: 
\code{\link{wlthhsps}}
}
\concept{Hispanics}
\concept{Intergroup Relations}
\keyword{variable}
