% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{workjews}
\alias{workjews}
\title{Hard working - lazy}
\source{
General Social Survey https://gss.norc.org
}
\description{
workjews
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
B. The second set of characteristics asks if people in the group tend to be hard-working or if they tend to be lazy.
2. Jews?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - hard-working
\item \verb{[7]} 7 - lazy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1185/vshow}{GSS Data Explorer page for \code{workjews}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 - hard-working \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - lazy \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 219 \tab 322 \tab 344 \tab 298 \tab 35 \tab 26 \tab 8 \tab 94 \tab 26 \tab 1372 \cr
   2000 \tab 1419 \tab 129 \tab 253 \tab 294 \tab 426 \tab 71 \tab 26 \tab 6 \tab 177 \tab 16 \tab 2817 \cr
   Total \tab 1419 \tab 348 \tab 575 \tab 638 \tab 724 \tab 106 \tab 52 \tab 14 \tab 271 \tab 42 \tab 4189 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthso}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workhsps}},
\code{\link{workso}},
\code{\link{workwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
