% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{workless}
\alias{workless}
\title{Cut back on your work?}
\source{
General Social Survey https://gss.norc.org
}
\description{
workless
}
\details{
Question In your (present/most recent) job, (have you ever done/did you ever do) any of the following because of your responsibilities to members of your family?
D. Cut back on your work?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1670/vshow}{GSS Data Explorer page for \code{workless}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not available in this year \tab don't know \tab skipped on web \tab Total \cr
   1996 \tab 1444 \tab 311 \tab 1034 \tab 115 \tab - \tab - \tab - \tab 2904 \cr
   2024 \tab 1162 \tab 561 \tab 1536 \tab 6 \tab - \tab 16 \tab 28 \tab 3309 \cr
   Total \tab 2606 \tab 872 \tab 2570 \tab 121 \tab 0 \tab 16 \tab 28 \tab 6213 \cr
}
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{earnmore}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{worklife}}

Other Family: 
\code{\link{balwkfam}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}}

Other Job: 
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}}

Other Work: 
\code{\link{balwkfam}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{worklife}}
}
\concept{Family}
\concept{Gender}
\concept{Job}
\concept{Work}
\keyword{variable}
