% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{workpol}
\alias{workpol}
\title{Has r worked for parties or candidates}
\source{
General Social Survey https://gss.norc.org
}
\description{
workpol
}
\details{
Question Have you done (other) work for one of the parties or candidates in most elections, some elections, only a few, or have you never done such work?
}
\section{Values}{

\itemize{
\item \verb{[1]} most elections
\item \verb{[2]} some elections
\item \verb{[3]} only a few
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1029/vshow}{GSS Data Explorer page for \code{workpol}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab most elections \tab some elections \tab only a few \tab never \tab no answer \tab Total \cr
   1987 \tab - \tab 48 \tab 166 \tab 251 \tab 1344 \tab 10 \tab 1819 \cr
   Total \tab 0 \tab 48 \tab 166 \tab 251 \tab 1344 \tab 10 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}}
}
\concept{Socio-political participation}
\keyword{variable}
