% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{workso}
\alias{workso}
\title{Hard working - lazy}
\source{
General Social Survey https://gss.norc.org
}
\description{
workso
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
B. The second set of characteristics asks if people in the group tend to be hard-working or if they tend to be lazy.
6. Southern Whites?
}
\section{Values}{

\itemize{
\item \code{1} 1 - hard-working
\item \code{7} 7 - lazy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1189/vshow}{GSS Data Explorer page for \code{workso}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 - hard-working \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - lazy \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 78 \tab 125 \tab 258 \tab 531 \tab 179 \tab 42 \tab 23 \tab 107 \tab 29 \tab 1372 \cr
   Total \tab 0 \tab 78 \tab 125 \tab 258 \tab 531 \tab 179 \tab 42 \tab 23 \tab 107 \tab 29 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthso}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
