% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{workweek}
\alias{workweek}
\title{Does r work at least 10 hrs/wk}
\source{
General Social Survey https://gss.norc.org
}
\description{
workweek
}
\details{
Question Do you usually work 10 hours or more per week for pay in your main job?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} not working
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4528/vshow}{GSS Data Explorer page for \code{workweek}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab yes \tab no \tab not working \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 898 \tab 160 \tab 395 \tab 84 \tab - \tab 1537 \cr
   Total \tab 0 \tab 898 \tab 160 \tab 395 \tab 84 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}}
}
\concept{Work Orientation}
\keyword{variable}
