% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_43.R
\name{world2}
\alias{world2}
\title{World image:the good must act vs. not}
\source{
General Social Survey https://gss.norc.org
}
\description{
world2
}
\details{
Question People have different images of the world and human nature. We'd like to know the kinds of images you have. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of the world  and human nature between the two contrasting images? (HAND CARD AH)
Look at the first set of contrasting images. If you think that "The world is basically filled with evil and sin," you would place yourself at 1. If you think "There is much goodness in the world which hints at God's goodness" you would place yourself at 7. If you think things are somewhere between these two you would place yourself at 2, 3,4, 5 or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F) Where would you place your image of the world on the scale for . . .?
(READ EACH SET AND CIRCLE ONE CODE FOR EACH)
}
\section{Values}{

\itemize{
\item \verb{[1]} the good must act
\item \verb{[7]} good must beware
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/367/vshow}{GSS Data Explorer page for \code{world2}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab the good must act \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab good must beware \tab don't know \tab no answer \tab Total \cr
   1985 \tab - \tab 341 \tab 197 \tab 243 \tab 415 \tab 121 \tab 61 \tab 127 \tab 9 \tab 20 \tab 1534 \cr
   Total \tab 0 \tab 341 \tab 197 \tab 243 \tab 415 \tab 121 \tab 61 \tab 127 \tab 9 \tab 20 \tab 1534 \cr
}
}

\seealso{
Other Core: 
\code{\link{contemp}},
\code{\link{libtemp}},
\code{\link{prottemp}},
\code{\link{world1}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}}
}
\concept{Core}
\keyword{variable}
