% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_43.R
\name{world3}
\alias{world3}
\title{World image:god not here vs. here}
\source{
General Social Survey https://gss.norc.org
}
\description{
world3
}
\details{
Question People have different images of the world and human nature. We'd like to know the kinds of images you have. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of the world  and human nature between the two contrasting images? (HAND CARD AH)
Look at the first set of contrasting images. If you think that "The world is basically filled with evil and sin," you would place yourself at 1. If you think "There is much goodness in the world which hints at God's goodness" you would place yourself at 7. If you think things are somewhere between these two you would place yourself at 2, 3,4, 5 or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F) Where would you place your image of the world on the scale for . . .?
(READ EACH SET AND CIRCLE ONE CODE FOR EACH)
}
\section{Values}{

\itemize{
\item \code{1} god not here
\item \code{7} god is here
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/368/vshow}{GSS Data Explorer page for \code{world3}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab god not here \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab god is here \tab don't know \tab no answer \tab Total \cr
   1985 \tab - \tab 61 \tab 33 \tab 39 \tab 297 \tab 279 \tab 237 \tab 530 \tab 23 \tab 35 \tab 1534 \cr
   Total \tab 0 \tab 61 \tab 33 \tab 39 \tab 297 \tab 279 \tab 237 \tab 530 \tab 23 \tab 35 \tab 1534 \cr
}
}

\seealso{
Other Core: 
\code{\link{contemp}},
\code{\link{libtemp}},
\code{\link{prottemp}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}}
}
\concept{Core}
\keyword{variable}
