% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_43.R
\name{world4}
\alias{world4}
\title{World image:man is good vs. corrupt}
\source{
General Social Survey https://gss.norc.org
}
\description{
world4
}
\details{
Question People have different images of the world and human nature. We'd like to know the kinds of images you have. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of the world  and human nature between the two contrasting images? (HAND CARD AH)
Look at the first set of contrasting images. If you think that "The world is basically filled with evil and sin," you would place yourself at 1. If you think "There is much goodness in the world which hints at God's goodness" you would place yourself at 7. If you think things are somewhere between these two you would place yourself at 2, 3,4, 5 or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F) Where would you place your image of the world on the scale for . . .?
(READ EACH SET AND CIRCLE ONE CODE FOR EACH)
}
\section{Values}{

\itemize{
\item \code{1} people are good
\item \code{7} people are evil
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989, 1990, 1994, 1996 \cr
   A/B/C \tab 1988, 1991, 1998 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/369/vshow}{GSS Data Explorer page for \code{world4}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab people are good \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab people are evil \tab don't know \tab no answer \tab Total \cr
   1985 \tab - \tab 417 \tab 316 \tab 269 \tab 264 \tab 86 \tab 46 \tab 104 \tab 6 \tab 26 \tab 1534 \cr
   1987 \tab - \tab 393 \tab 314 \tab 319 \tab 429 \tab 139 \tab 50 \tab 141 \tab - \tab 34 \tab 1819 \cr
   1988 \tab - \tab 310 \tab 301 \tab 255 \tab 370 \tab 92 \tab 43 \tab 87 \tab 5 \tab 18 \tab 1481 \cr
   1989 \tab 531 \tab 216 \tab 217 \tab 179 \tab 235 \tab 57 \tab 31 \tab 53 \tab 1 \tab 17 \tab 1537 \cr
   1990 \tab 444 \tab 165 \tab 193 \tab 169 \tab 262 \tab 63 \tab 34 \tab 32 \tab 2 \tab 8 \tab 1372 \cr
   1991 \tab 158 \tab 194 \tab 277 \tab 257 \tab 298 \tab 120 \tab 60 \tab 70 \tab 2 \tab 81 \tab 1517 \cr
   1994 \tab 1522 \tab 233 \tab 319 \tab 265 \tab 388 \tab 122 \tab 43 \tab 77 \tab 4 \tab 19 \tab 2992 \cr
   1996 \tab 1913 \tab 168 \tab 182 \tab 199 \tab 243 \tab 91 \tab 35 \tab 65 \tab 2 \tab 6 \tab 2904 \cr
   1998 \tab 1387 \tab 213 \tab 295 \tab 279 \tab 368 \tab 123 \tab 57 \tab 91 \tab 5 \tab 14 \tab 2832 \cr
   Total \tab 5955 \tab 2309 \tab 2414 \tab 2191 \tab 2857 \tab 893 \tab 399 \tab 720 \tab 27 \tab 223 \tab 17988 \cr
}
}

\seealso{
Other Core: 
\code{\link{contemp}},
\code{\link{libtemp}},
\code{\link{prottemp}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}}
}
\concept{Core}
\keyword{variable}
