% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_43.R
\name{world6}
\alias{world6}
\title{World image:chaos vs. harmony}
\source{
General Social Survey https://gss.norc.org
}
\description{
world6
}
\details{
Question People have different images of the world and human nature. We'd like to know the kinds of images you have. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of the world  and human nature between the two contrasting images? (HAND CARD AH)
Look at the first set of contrasting images. If you think that "The world is basically filled with evil and sin," you would place yourself at 1. If you think "There is much goodness in the world which hints at God's goodness" you would place yourself at 7. If you think things are somewhere between these two you would place yourself at 2, 3,4, 5 or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F) Where would you place your image of the world on the scale for . . .?
(READ EACH SET AND CIRCLE ONE CODE FOR EACH)
}
\section{Values}{

\itemize{
\item \code{1} world is chaotic
\item \code{7} world is harmonious
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/371/vshow}{GSS Data Explorer page for \code{world6}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab world is chaotic \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab world is harmonious \tab don't know \tab no answer \tab Total \cr
   1985 \tab - \tab 163 \tab 128 \tab 215 \tab 565 \tab 242 \tab 83 \tab 102 \tab 13 \tab 23 \tab 1534 \cr
   Total \tab 0 \tab 163 \tab 128 \tab 215 \tab 565 \tab 242 \tab 83 \tab 102 \tab 13 \tab 23 \tab 1534 \cr
}
}

\seealso{
Other Core: 
\code{\link{contemp}},
\code{\link{libtemp}},
\code{\link{prottemp}},
\code{\link{world1}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world7}}
}
\concept{Core}
\keyword{variable}
