% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{worry}
\alias{worry}
\title{How often could not control worrying in past 2 weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
worry
}
\details{
Question Over the last 2 weeks, how often have you been bothered by the following problems: Not being able to stop or control worrying.
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[2]} several days
\item \verb{[3]} more than half the days
\item \verb{[4]} nearly every day
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7860/vshow}{GSS Data Explorer page for \code{worry}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all \tab several days \tab more than half the days \tab nearly every day \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1557 \tab 1114 \tab 554 \tab 134 \tab 137 \tab 7 \tab 20 \tab 21 \tab 3544 \cr
   Total \tab 1557 \tab 1114 \tab 554 \tab 134 \tab 137 \tab 7 \tab 20 \tab 21 \tab 3544 \cr
}
}

\keyword{variable}
