% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{worship}
\alias{worship}
\title{How good local church wor'ship services}
\source{
General Social Survey https://gss.norc.org
}
\description{
worship
}
\details{
Question How do you rate how your local church does each of the following?  Would you say your local church does an excellent, very good, fair, or poor job of:
E. Worship services
}
\section{Values}{

\itemize{
\item \verb{[1]} excellent
\item \verb{[2]} very good
\item \verb{[3]} good
\item \verb{[4]} fair
\item \verb{[5]} poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1165/vshow}{GSS Data Explorer page for \code{worship}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1831 \tab 298 \tab 286 \tab 224 \tab 47 \tab 7 \tab 50 \tab 74 \tab 2817 \cr
   Total \tab 1831 \tab 298 \tab 286 \tab 224 \tab 47 \tab 7 \tab 50 \tab 74 \tab 2817 \cr
}
}

\seealso{
Other Religion: 
\code{\link{ferespct}},
\code{\link{obeytch}},
\code{\link{preach}},
\code{\link{sympcoun}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
