% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_478.R
\name{wrkbaby}
\alias{wrkbaby}
\title{Shld woman with preschooler work?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkbaby
}
\details{
Question Do you think that women should work outside the home full-time, part-time or not at all under these circumstances:
B. When there is a child under school age.
}
\section{Values}{

\itemize{
\item \code{1} work full-time
\item \code{2} work part-time
\item \code{3} stay home
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2022 \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4354/vshow}{GSS Data Explorer page for \code{wrkbaby}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab work full-time \tab work part-time \tab stay home \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1988 \tab 67 \tab 136 \tab 425 \tab 681 \tab 143 \tab 29 \tab - \tab 1481 \cr
   1994 \tab 1545 \tab 141 \tab 436 \tab 696 \tab 140 \tab 34 \tab - \tab 2992 \cr
   2002 \tab 1594 \tab 133 \tab 421 \tab 460 \tab 155 \tab 2 \tab - \tab 2765 \cr
   2012 \tab 672 \tab 171 \tab 481 \tab 335 \tab 306 \tab 9 \tab - \tab 1974 \cr
   2022 \tab 2380 \tab 209 \tab 315 \tab 212 \tab 411 \tab 9 \tab 8 \tab 3544 \cr
   Total \tab 6258 \tab 790 \tab 2078 \tab 2384 \tab 1155 \tab 83 \tab 8 \tab 12756 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{ilikejob}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
