% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{wrkearn}
\alias{wrkearn}
\title{Job is just a way to earn money}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkearn
}
\details{
Question Thinking of work in general, please circle one number for each statement below to show how much you agree or disagree with each.
A. A job is just a way of earning money - no more
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998, 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4492/vshow}{GSS Data Explorer page for \code{wrkearn}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 104 \tab 215 \tab 251 \tab 569 \tab 279 \tab 15 \tab 104 \tab - \tab 1537 \cr
   1998 \tab - \tab 75 \tab 209 \tab 198 \tab 483 \tab 201 \tab 25 \tab 1641 \tab - \tab 2832 \cr
   2006 \tab - \tab 124 \tab 262 \tab 213 \tab 655 \tab 257 \tab 6 \tab 1 \tab 2992 \tab 4510 \cr
   2016 \tab - \tab 128 \tab 277 \tab 242 \tab 605 \tab 220 \tab 1 \tab 1394 \tab - \tab 2867 \cr
   Total \tab 0 \tab 431 \tab 963 \tab 904 \tab 2312 \tab 957 \tab 47 \tab 3140 \tab 2992 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{hwduties}},
\code{\link{paysame}},
\code{\link{sexemp}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}},
\code{\link{yrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
