% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{wrkenjoy}
\alias{wrkenjoy}
\title{Enjoy job even if $ not needed}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkenjoy
}
\details{
Question Thinking of work in general, please circle one number for each statement below to show how much you agree or disagree with each.
B. I would enjoy having a paying job even if I did not need that money
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4493/vshow}{GSS Data Explorer page for \code{wrkenjoy}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 169 \tab 702 \tab 227 \tab 230 \tab 64 \tab 37 \tab 108 \tab - \tab 1537 \cr
   1998 \tab - \tab 112 \tab 583 \tab 186 \tab 206 \tab 75 \tab 34 \tab 1636 \tab - \tab 2832 \cr
   2006 \tab - \tab 214 \tab 750 \tab 177 \tab 266 \tab 103 \tab 6 \tab 2 \tab 2992 \tab 4510 \cr
   2016 \tab - \tab 230 \tab 796 \tab 178 \tab 192 \tab 74 \tab 2 \tab 5 \tab 1390 \tab 2867 \cr
   Total \tab 0 \tab 725 \tab 2831 \tab 768 \tab 894 \tab 316 \tab 79 \tab 1751 \tab 4382 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{hwduties}},
\code{\link{paysame}},
\code{\link{sexemp}},
\code{\link{wrkearn}},
\code{\link{wrkimp}},
\code{\link{yrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
