% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_478.R
\name{wrkgrown}
\alias{wrkgrown}
\title{Shld woman work after kids leave home?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkgrown
}
\details{
Question Do you think that women should work outside the home full-time, part-time or not at all under these circumstances:
D. After the children leave home.
}
\section{Values}{

\itemize{
\item \verb{[1]} work full-time
\item \verb{[2]} work part-time
\item \verb{[3]} stay home
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4356/vshow}{GSS Data Explorer page for \code{wrkgrown}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab work full-time \tab work part-time \tab stay home \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 888 \tab 236 \tab 62 \tab 197 \tab 31 \tab 1481 \cr
   1994 \tab 1545 \tab 986 \tab 203 \tab 30 \tab 200 \tab 28 \tab 2992 \cr
   2002 \tab 1594 \tab 735 \tab 203 \tab 28 \tab 203 \tab 2 \tab 2765 \cr
   Total \tab 3206 \tab 2609 \tab 642 \tab 120 \tab 600 \tab 61 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{ilikejob}},
\code{\link{wrkbaby}},
\code{\link{wrknokid}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
