% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{wrkhome}
\alias{wrkhome}
\title{How often r works at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkhome
}
\details{
Question How often do you work at home as part of your job?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} a few times a year
\item \code{3} about once a month
\item \code{4} about once a week
\item \code{5} more than once a week
\item \code{6} worker works mainly at home
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2767/vshow}{GSS Data Explorer page for \code{wrkhome}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{hrsrelax}},
\code{\link{learnnew}},
\code{\link{secondwk}},
\code{\link{whywkhme}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}}
}
\concept{Quality of Working Life}
\keyword{variable}
