% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_284.R
\name{wrkhro}
\alias{wrkhro}
\title{Hours per week using other computer make money}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkhro
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
C. Paid work or other income-producing activities?
2. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2590/vshow}{GSS Data Explorer page for \code{wrkhro}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 40 \tab 6 \tab 60 \tab no answer \tab Total \cr
   2000 \tab 2665 \tab 85 \tab 7 \tab 4 \tab 1 \tab 1 \tab 1 \tab 53 \tab 2817 \cr
   Total \tab 2665 \tab 85 \tab 7 \tab 4 \tab 1 \tab 1 \tab 1 \tab 53 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emhro}},
\code{\link{emmino}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{wrkmino}},
\code{\link{wwwhro}},
\code{\link{wwwmino}}
}
\concept{Information Society}
\keyword{variable}
