% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{wrkimp}
\alias{wrkimp}
\title{Work is persons most impt. activity}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkimp
}
\details{
Question Thinking of work in general, please circle one number for each statement below to show how much you agree or disagree with each.
C. Work is a person's most important activity
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4494/vshow}{GSS Data Explorer page for \code{wrkimp}}.
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{hwduties}},
\code{\link{paysame}},
\code{\link{sexemp}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{yrsfirm}}

Other Work: 
\code{\link{wrkearn}},
\code{\link{wrkenjoy}}
}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
