% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{wrkindep}
\alias{wrkindep}
\title{I can work independently}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkindep
}
\details{
Question How true would you say each of the following statements is about your job?  Would you say very true, true, somewhat true, or not true at all:
A.I can work independently.
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} true
\item \verb{[3]} somewhat true
\item \verb{[4]} not true at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1308/vshow}{GSS Data Explorer page for \code{wrkindep}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab true \tab somewhat true \tab not true at all \tab can't choose \tab no answer \tab not available in this year \tab Total \cr
   1991 \tab 605 \tab 414 \tab 302 \tab 111 \tab 73 \tab 2 \tab 10 \tab - \tab 1517 \cr
   Total \tab 605 \tab 414 \tab 302 \tab 111 \tab 73 \tab 2 \tab 10 \tab 0 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{afailure}},
\code{\link{automatn}},
\code{\link{baseofop}},
\code{\link{chngeorg}},
\code{\link{cowrkrel}},
\code{\link{drugtest}},
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{futpromo}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{helporg}},
\code{\link{idecide}},
\code{\link{imatter}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobcntrl}},
\code{\link{jobjudge}},
\code{\link{jobvshme}},
\code{\link{jobyears}},
\code{\link{learning}},
\code{\link{levels}()},
\code{\link{lottosay}},
\code{\link{ltrsref}},
\code{\link{merged}},
\code{\link{monitred}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{offsup}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{orgfin}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{othpay}},
\code{\link{outofbiz}},
\code{\link{pessimst}},
\code{\link{physical}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{proudorg}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{racpromo}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{reorg}},
\code{\link{samejob}},
\code{\link{samevals}},
\code{\link{satself}},
\code{\link{schoolng}},
\code{\link{sexpromo}},
\code{\link{skiltest}},
\code{\link{slfrspct}},
\code{\link{splocnum}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{totunemp}},
\code{\link{training}},
\code{\link{unmanrel}},
\code{\link{unpower}},
\code{\link{unprog}},
\code{\link{unvote}},
\code{\link{wherewrk}},
\code{\link{wrkmuch}},
\code{\link{wrkwell}},
\code{\link{wrkyears}},
\code{\link{youmoney}},
\code{\link{yousup}}

Other Work: 
\code{\link{absent}},
\code{\link{accntsci}},
\code{\link{architct}},
\code{\link{automatn}},
\code{\link{autonojb}},
\code{\link{balwkfam}},
\code{\link{buyvalue}},
\code{\link{chngwork}},
\code{\link{cmprgmng}},
\code{\link{commute}},
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{denyrais}},
\code{\link{edcskill}},
\code{\link{eharaswk}},
\code{\link{empinput}},
\code{\link{empself}},
\code{\link{emptrain}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{esop}},
\code{\link{evharjb}},
\code{\link{extr2005}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkoff}},
\code{\link{farming}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{flexhrs}},
\code{\link{flextime}},
\code{\link{getahead}},
\code{\link{getaheay}},
\code{\link{harjob5}},
\code{\link{harsexjb}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{idecide}},
\code{\link{ignorwk}},
\code{\link{impjob}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jobcntrl}},
\code{\link{jobfind}},
\code{\link{jobfind1}},
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobinc}},
\code{\link{jobjudge}},
\code{\link{jobkeep}},
\code{\link{joblose}},
\code{\link{jobmeans}},
\code{\link{jobpay}},
\code{\link{jobpromo}},
\code{\link{jobrise}},
\code{\link{jobsat}},
\code{\link{jobsec}},
\code{\link{jobskill}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{jokeswk}},
\code{\link{journlsm}},
\code{\link{kdwrkslf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{lawenfrc}},
\code{\link{layoffs}},
\code{\link{leisure}},
\code{\link{levels}()},
\code{\link{liedcwkr}},
\code{\link{lottosay}},
\code{\link{marrcoun}},
\code{\link{mawrkslf}},
\code{\link{medtreat}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{monitred}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{negjob5}},
\code{\link{newjobs}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{numorg}},
\code{\link{offsup}},
\code{\link{ops2005}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{parleave}},
\code{\link{pawrkslf}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{phyeffrt}},
\code{\link{placewrk}},
\code{\link{prideorg}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{richwork}},
\code{\link{rimpskls}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{satjob}},
\code{\link{satjob7}},
\code{\link{sbwrkslf}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientre}},
\code{\link{scientwk}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexhar}},
\code{\link{skltrain}},
\code{\link{slfmangd}},
\code{\link{slsmnshp}},
\code{\link{spdays}},
\code{\link{sphedslf}},
\code{\link{sphour}},
\code{\link{spvtrfair}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{spwrkslf}},
\code{\link{stayorg3}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{talkteam}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{toldwork}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{trynewjb}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{useskill}},
\code{\link{wantjob1}},
\code{\link{waypaid}},
\code{\link{wealth}},
\code{\link{weekswrk}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkcontct}},
\code{\link{wkfambiz}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkracism}},
\code{\link{wkrspns}},
\code{\link{wksexism}},
\code{\link{wksub}},
\code{\link{wksubs}},
\code{\link{wksup}},
\code{\link{wksups}},
\code{\link{wktenure}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{workdy}},
\code{\link{workfor}},
\code{\link{workhr}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{worknow}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkhome}},
\code{\link{wrkimp}},
\code{\link{wrkindp}},
\code{\link{wrkmuch}},
\code{\link{wrksched}},
\code{\link{wrkslf}},
\code{\link{wrkslfst}},
\code{\link{wrkwell}},
\code{\link{yearsjob}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Work}
\concept{Work Organizations}
\keyword{variable}
