% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_284.R
\name{wrkmino}
\alias{wrkmino}
\title{Minutes per week using other computer make money}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkmino
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
C. Paid work or other income-producing activities?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2589/vshow}{GSS Data Explorer page for \code{wrkmino}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 15 \tab 30 \tab 5 \tab no answer \tab Total \cr
   2000 \tab 2665 \tab 111 \tab 1 \tab 1 \tab 2 \tab 1 \tab 36 \tab 2817 \cr
   Total \tab 2665 \tab 111 \tab 1 \tab 1 \tab 2 \tab 1 \tab 36 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emhro}},
\code{\link{emmino}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{wrkhro}},
\code{\link{wwwhro}},
\code{\link{wwwmino}}
}
\concept{Information Society}
\keyword{variable}
