% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_143.R
\name{wrkmuch}
\alias{wrkmuch}
\title{How much you compare to people do same job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkmuch
}
\details{
Question Compared to other people who do the same or similar kind of work that you do, how much work would you say you do?  Would you say that you do much more, somewhat more, about the same, somewhat less or much less?
}
\section{Values}{

\itemize{
\item \verb{[1]} much more
\item \verb{[2]} somewhat more
\item \verb{[3]} about same
\item \verb{[4]} somewhat less
\item \verb{[5]} much less
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1339/vshow}{GSS Data Explorer page for \code{wrkmuch}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much more \tab somewhat more \tab about same \tab somewhat less \tab much less \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 203 \tab 338 \tab 317 \tab 28 \tab 2 \tab 14 \tab 10 \tab 1517 \cr
   Total \tab 605 \tab 203 \tab 338 \tab 317 \tab 28 \tab 2 \tab 14 \tab 10 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{chngeorg}},
\code{\link{helporg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{samevals}},
\code{\link{stayorg1}},
\code{\link{stayorg2}}
}
\concept{Work Organizations}
\keyword{variable}
