% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_478.R
\name{wrknokid}
\alias{wrknokid}
\title{Shld mom work aftr marriage, before kids?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrknokid
}
\details{
Question Do you think that women should work outside the home full-time, part-time or not at all under these circumstances:
A. After marrying and before there are children.
}
\section{Values}{

\itemize{
\item \code{1} work full-time
\item \code{2} work part-time
\item \code{3} stay home
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4353/vshow}{GSS Data Explorer page for \code{wrknokid}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab work full-time \tab work part-time \tab stay home \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 931 \tab 215 \tab 67 \tab 171 \tab 30 \tab 1481 \cr
   1994 \tab 1545 \tab 1067 \tab 168 \tab 34 \tab 154 \tab 24 \tab 2992 \cr
   2002 \tab 1594 \tab 796 \tab 168 \tab 20 \tab 185 \tab 2 \tab 2765 \cr
   Total \tab 3206 \tab 2794 \tab 551 \tab 121 \tab 510 \tab 56 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{ilikejob}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
