% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{wrkoth}
\alias{wrkoth}
\title{Ever worked full time at other parts of org}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkoth
}
\details{
Question A. Have you ever worked full time at other parts (locations/schools/government departments or agencies) of the larger organization?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2141/vshow}{GSS Data Explorer page for \code{wrkoth}}.
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}}
}
\concept{Jobs}
\keyword{variable}
