% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_01.R
\name{wrkslf}
\alias{wrkslf}
\title{R self-emp or works for somebody}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkslf
}
\details{
Question A. What kind of work do you (did you normally) do?  That is, what (is/was) your job called?
OCCUPATION:  (See REMARKS)
B. What (do/did) you actually do in that job?  Tell me, what (are/were) some of your main duties?
C. What kind of place (do/did) you work for?
INDUSTRY:  (See REMARKS)
D. What (do/did) they (make/do)?
E. (Are/Were) you self employed or (do/did) you work for someone else?
F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)?
G. About how much time (does/did) it usually take you to travel to work - about how many minutes?
}
\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/9/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no answer \tab self-employed \tab someone else \tab don't know \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 162 \tab 3 \tab 149 \tab 1299 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 147 \tab 10 \tab 148 \tab 1199 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 128 \tab 2 \tab 116 \tab 1238 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 134 \tab 8 \tab 143 \tab 1205 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 145 \tab 2 \tab 137 \tab 1215 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 113 \tab - \tab 131 \tab 1286 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 127 \tab 1 \tab 140 \tab 1264 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 119 \tab 7 \tab 168 \tab 1174 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 122 \tab 7 \tab 186 \tab 1545 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 104 \tab 7 \tab 168 \tab 1320 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 88 \tab 5 \tab 170 \tab 1210 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 81 \tab - \tab 193 \tab 1260 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 100 \tab - \tab 187 \tab 1183 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 95 \tab 6 \tab 187 \tab 1531 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 85 \tab 6 \tab 163 \tab 1227 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 93 \tab 25 \tab 171 \tab 1248 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 71 \tab 2 \tab 150 \tab 1149 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 97 \tab 5 \tab 157 \tab 1258 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 79 \tab 9 \tab 187 \tab 1331 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 126 \tab 19 \tab 339 \tab 2508 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 107 \tab 45 \tab 365 \tab 2387 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 135 \tab 42 \tab 323 \tab 2332 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 125 \tab 54 \tab 307 \tab 2331 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 91 \tab 2 \tab 307 \tab 2362 \tab 3 \tab - \tab - \tab 2765 \cr
   2004 \tab 113 \tab 1 \tab 343 \tab 2353 \tab 2 \tab - \tab - \tab 2812 \cr
   2006 \tab 187 \tab 13 \tab 508 \tab 3799 \tab 3 \tab - \tab - \tab 4510 \cr
   2008 \tab 62 \tab 3 \tab 226 \tab 1732 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 98 \tab 4 \tab 234 \tab 1706 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 78 \tab 1 \tab 194 \tab 1700 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab 92 \tab 4 \tab 295 \tab 2145 \tab 2 \tab - \tab - \tab 2538 \cr
   2016 \tab 90 \tab 5 \tab 296 \tab 2472 \tab 4 \tab - \tab - \tab 2867 \cr
   2018 \tab 77 \tab 7 \tab 232 \tab 2029 \tab 3 \tab - \tab - \tab 2348 \cr
   2021 \tab 79 \tab 5 \tab 432 \tab 3493 \tab 2 \tab 21 \tab - \tab 4032 \cr
   2022 \tab 113 \tab 7 \tab 393 \tab 3013 \tab 3 \tab 15 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 3663 \tab 317 \tab 7845 \tab 60504 \tab 25 \tab 36 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{evwork}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{id}},
\code{\link{occ}},
\code{\link{prestige}},
\code{\link{wrkgovt}},
\code{\link{wrkstat}},
\code{\link{year}}

Other Occupation: 
\code{\link{occ}},
\code{\link{prestige}},
\code{\link{wrkgovt}}

Other Work: 
\code{\link{hrs1}},
\code{\link{hrs2}}
}
\concept{Core}
\concept{Job}
\concept{Occupation}
\concept{Work}
\keyword{variable}
