% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{wrkslfst}
\alias{wrkslfst}
\title{R 1st job: self-employed or no?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkslfst
}
\details{
Question Respondent's First Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item \verb{[1]} self employed
\item \verb{[2]} someone else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1464/vshow}{GSS Data Explorer page for \code{wrkslfst}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab self employed \tab someone else \tab no answer \tab Total \cr
   1994 \tab 1587 \tab 63 \tab 1317 \tab 25 \tab 2992 \cr
   Total \tab 1587 \tab 63 \tab 1317 \tab 25 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{indfirst}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{payrdied}},
\code{\link{presfrst}}
}
\concept{Family Mobility}
\keyword{variable}
