% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{wrktime}
\alias{wrktime}
\title{R has enough time to get the job done}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrktime
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
M. I have enough time to get the job done
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2808/vshow}{GSS Data Explorer page for \code{wrktime}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 801 \tab 665 \tab 193 \tab 115 \tab 7 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 687 \tab 723 \tab 205 \tab 86 \tab 15 \tab 18 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 482 \tab 491 \tab 126 \tab 59 \tab 4 \tab 25 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 537 \tab 524 \tab 142 \tab 38 \tab 1 \tab 8 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 615 \tab 567 \tab 167 \tab 60 \tab 5 \tab 34 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 808 \tab 833 \tab 211 \tab 94 \tab 8 \tab 14 \tab 21 \tab 3544 \cr
   Total \tab 8345 \tab 3930 \tab 3803 \tab 1044 \tab 452 \tab 40 \tab 114 \tab 21 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}}
}
\concept{Quality of Working Life}
\keyword{variable}
