% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_656.R
\name{wrkwayup_next}
\alias{wrkwayup_next}
\title{Black people overcome prejudice without favors - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkwayup_next
}
\details{
Question Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or disagree strongly with the following statements? Irish, Italian, Jewish and many other minorities overcame prejudice and worked their way up. Black people should do

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree somewhat
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7892/vshow}{GSS Data Explorer page for \code{wrkwayup_next}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab neither agree nor disagree \tab disagree somewhat \tab disagree strongly \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 276 \tab 269 \tab 223 \tab 207 \tab 259 \tab 1 \tab 3544 \cr
   Total \tab 2309 \tab 276 \tab 269 \tab 223 \tab 207 \tab 259 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
