% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_142.R
\name{wrkwell}
\alias{wrkwell}
\title{How well youcompare to people do same job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkwell
}
\details{
Question Compared to other people who do the same or similar kind of work that you do, how well would you say you do your job?  Would you say much better, somewhat better, about the same, somewhat worse or much worse?
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} somewhat better
\item \verb{[3]} about same
\item \verb{[4]} somewhat worse
\item \verb{[5]} much worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1338/vshow}{GSS Data Explorer page for \code{wrkwell}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab much better \tab somewhat better \tab about same \tab somewhat worse \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 299 \tab 343 \tab 248 \tab 2 \tab 11 \tab 9 \tab 1517 \cr
   Total \tab 605 \tab 299 \tab 343 \tab 248 \tab 2 \tab 11 \tab 9 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{totmoney}},
\code{\link{youmoney}}
}
\concept{Work Organizations}
\keyword{variable}
