% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{wrthless}
\alias{wrthless}
\title{In past 30 days feel worthless}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrthless
}
\details{
Question In the past 30 days, about how often did you feel:
F. Worthless.
}
\section{Values}{

\itemize{
\item \code{1} all of the time
\item \code{2} most of the time
\item \code{3} some of the time
\item \code{4} little of time
\item \code{5} none of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2119/vshow}{GSS Data Explorer page for \code{wrthless}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{beausprt}},
\code{\link{effort}},
\code{\link{fidgety}},
\code{\link{godlove}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{uniongod}}
}
\concept{Religion and Culture}
\keyword{variable}
