% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{wtrpollu}
\alias{wtrpollu}
\title{Neighborhood affected by water pollution over the last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
wtrpollu
}
\details{
Question To what degree was your neighborhood affected by water polution over the last twelve months?
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all
\item \verb{[2]} to a small extent
\item \verb{[3]} to some extent
\item \verb{[4]} to a great extent
\item \verb{[5]} to a very great extent
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7502/vshow}{GSS Data Explorer page for \code{wtrpollu}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab not at all \tab to a small extent \tab to some extent \tab to a great extent \tab to a very great extent \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 854 \tab 449 \tab 285 \tab 88 \tab 23 \tab 110 \tab 201 \tab 19 \tab 4032 \cr
   Total \tab 2003 \tab 854 \tab 449 \tab 285 \tab 88 \tab 23 \tab 110 \tab 201 \tab 19 \tab 4032 \cr
}
}

\keyword{variable}
