% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_611.R
\name{wtss}
\alias{wtss}
\title{Weight variable}
\source{
General Social Survey https://gss.norc.org
}
\description{
wtss
}
\details{
Question Weight of variables.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{vpsu}},
\code{\link{vstrat}},
\code{\link{wtssall}},
\code{\link{wtssnr}}

Other Sample: 
\code{\link{vpsu}},
\code{\link{wtssall}},
\code{\link{wtssnr}}
}
\concept{Sample}
\concept{Sampling and Weights}
\keyword{variable}
