% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{wtsscompnr}
\alias{wtsscompnr}
\title{Wtsscompnr}
\source{
General Social Survey https://gss.norc.org
}
\description{
wtsscompnr
}
\details{
Question
}
\section{Values}{

\itemize{
\item \verb{[Empty]} No applicable value labels
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\keyword{variable}
