% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{wtssnrps}
\alias{wtssnrps}
\title{Person post-stratification weight, nonrespondents adjusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
wtssnrps
}
\details{
Question Person post-stratification weight, nonrespondents adjusted
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
}
}

\keyword{variable}
