% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{wtssnrps_nea}
\alias{wtssnrps_nea}
\title{Poststratification and nonresponse weight for analyzing nea follow-on}
\source{
General Social Survey https://gss.norc.org
}
\description{
wtssnrps_nea
}
\details{
Question POSTSTRATIFICATION AND NONRESPONSE WEIGHT FOR ANALYZING NEA FOLLOW-ON
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\keyword{variable}
