% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{wwwart}
\alias{wwwart}
\title{Use web to learn about visual arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwart
}
\details{
Question A. Do you ever use the Web to learn about, view, or discuss topics related to the visual arts–painting, sculpture, or so on?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2473/vshow}{GSS Data Explorer page for \code{wwwart}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2326 \tab 100 \tab 278 \tab 113 \tab 2817 \cr
   2002 \tab 2284 \tab 91 \tab 344 \tab 46 \tab 2765 \cr
   Total \tab 4610 \tab 191 \tab 622 \tab 159 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{musicnew}},
\code{\link{wwwworld}}
}
\concept{Information Society}
\keyword{variable}
