% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_271.R
\name{wwwcntry}
\alias{wwwcntry}
\title{In last month listen country music on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwcntry
}
\details{
Question In the last month, how many times have you used the Web to learn about or listen to the following kinds of music?
B. Country?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} more than 2 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2462/vshow}{GSS Data Explorer page for \code{wwwcntry}}.
}

\seealso{
Other Information Society: 
\code{\link{musicbuy}},
\code{\link{musiclst}},
\code{\link{wwwclass}},
\code{\link{wwwgthic}},
\code{\link{wwwjazz}},
\code{\link{wwwoldie}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}}

Other Computers And Internet: 
\code{\link{musicbuy}},
\code{\link{musiclst}},
\code{\link{wwwclass}},
\code{\link{wwwgthic}},
\code{\link{wwwjazz}},
\code{\link{wwwoldie}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
