% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{wwwhome}
\alias{wwwhome}
\title{Can r use www at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwhome
}
\details{
Question A. Can you use the World Wide Web when you are at home?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2552/vshow}{GSS Data Explorer page for \code{wwwhome}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2418 \tab 323 \tab 53 \tab 23 \tab 2817 \cr
   2002 \tab 2016 \tab 665 \tab 82 \tab 2 \tab 2765 \cr
   Total \tab 4434 \tab 988 \tab 135 \tab 25 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwline1}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
