% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{wwwhr}
\alias{wwwhr}
\title{Www hours per week}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwhr
}
\details{
Question Not counting e-mail, about how many minutes or hours per week Do you use the Web? (Include time you spend visiting regular web sites and time spent using interactive Internet services like chat rooms, Usenet groups, discussion forums, bulletin boards, and the like.)
A. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[0]} 0 hours
\item \verb{[168]} 168 hours
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C \tab 2000, 2002, 2004, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2384/vshow}{GSS Data Explorer page for \code{wwwhr}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 hours \tab 1 \tab 10 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 2 \tab 20 \tab 21 \tab 23 \tab 25 \tab 28 \tab 3 \tab 30 \tab 35 \tab 36 \tab 4 \tab 40 \tab 42 \tab 45 \tab 46 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab 98 \tab don't know \tab no answer \tab 100 \tab 11 \tab 112 \tab 18 \tab 22 \tab 24 \tab 26 \tab 27 \tab 29 \tab 32 \tab 37 \tab 48 \tab 64 \tab 130 \tab 49 \tab 55 \tab 65 \tab 70 \tab 72 \tab 75 \tab 80 \tab 120 \tab 168 hours \tab 38 \tab 66 \tab 19 \tab 33 \tab 63 \tab 84 \tab 122 \tab 90 \tab 91 \tab 132 \tab 140 \tab 34 \tab 43 \tab 47 \tab 56 \tab 85 \tab 86 \tab 105 \tab 110 \tab 125 \tab 150 \tab 160 \tab 39 \tab 57 \tab 68 \tab 83 \tab 95 \tab skipped on web \tab 119 \tab 126 \tab 41 \tab 51 \tab 54 \tab 76 \tab 96 \tab not available in this year \tab Total \cr
   2000 \tab 1763 \tab 202 \tab 163 \tab 67 \tab 9 \tab 1 \tab 10 \tab 21 \tab 4 \tab 1 \tab 143 \tab 26 \tab 3 \tab 1 \tab 9 \tab 1 \tab 75 \tab 9 \tab 1 \tab 1 \tab 60 \tab 8 \tab 1 \tab 2 \tab 1 \tab 82 \tab 1 \tab 42 \tab 1 \tab 31 \tab 17 \tab 4 \tab 1 \tab 1 \tab 55 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1116 \tab 249 \tab 279 \tab 110 \tab 23 \tab 2 \tab 14 \tab 44 \tab 3 \tab 2 \tab 220 \tab 49 \tab 6 \tab 1 \tab 8 \tab 1 \tab 129 \tab 16 \tab 8 \tab - \tab 92 \tab 12 \tab - \tab 2 \tab - \tab 135 \tab 5 \tab 46 \tab 2 \tab 45 \tab 41 \tab 5 \tab - \tab 5 \tab 70 \tab 1 \tab 4 \tab 1 \tab 2 \tab 3 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 978 \tab 172 \tab 283 \tab 147 \tab 24 \tab 4 \tab 32 \tab 47 \tab 5 \tab 1 \tab 233 \tab 72 \tab 9 \tab 2 \tab 21 \tab 3 \tab 119 \tab 24 \tab 11 \tab 1 \tab 99 \tab 22 \tab - \tab 4 \tab - \tab 155 \tab 8 \tab 53 \tab 4 \tab 73 \tab 45 \tab 6 \tab - \tab 14 \tab 119 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2505 \tab 142 \tab 264 \tab 171 \tab 36 \tab 4 \tab 33 \tab 72 \tab 4 \tab 1 \tab 220 \tab 87 \tab 9 \tab 1 \tab 33 \tab 8 \tab 149 \tab 51 \tab 12 \tab - \tab 98 \tab 29 \tab 3 \tab 2 \tab - \tab 156 \tab 12 \tab 54 \tab 7 \tab 69 \tab 32 \tab 5 \tab - \tab 16 \tab 198 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 7 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2010 \tab 988 \tab 90 \tab 131 \tab 83 \tab 14 \tab - \tab 32 \tab 43 \tab 6 \tab 1 \tab 124 \tab 59 \tab 5 \tab 2 \tab 17 \tab 4 \tab 80 \tab 27 \tab 8 \tab 1 \tab 56 \tab 20 \tab 1 \tab 5 \tab - \tab 80 \tab 10 \tab 35 \tab 3 \tab 48 \tab 40 \tab 3 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 953 \tab 89 \tab 112 \tab 81 \tab 19 \tab 3 \tab 26 \tab 32 \tab 2 \tab - \tab 103 \tab 58 \tab 10 \tab - \tab 20 \tab 9 \tab 83 \tab 29 \tab 6 \tab 2 \tab 55 \tab 21 \tab 1 \tab 5 \tab - \tab 90 \tab 8 \tab 43 \tab 6 \tab 48 \tab 29 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1126 \tab 121 \tab 142 \tab 128 \tab 42 \tab 2 \tab 35 \tab 64 \tab 6 \tab 2 \tab 146 \tab 91 \tab 13 \tab - \tab 28 \tab 10 \tab 92 \tab 41 \tab 13 \tab 4 \tab 60 \tab 33 \tab 1 \tab 6 \tab - \tab 104 \tab 15 \tab 41 \tab 12 \tab 59 \tab 39 \tab 6 \tab - \tab 10 \tab 3 \tab 4 \tab 1 \tab - \tab 6 \tab 1 \tab 8 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1274 \tab 137 \tab 143 \tab 159 \tab 34 \tab 2 \tab 49 \tab 97 \tab 5 \tab 3 \tab 115 \tab 117 \tab 23 \tab - \tab 37 \tab 9 \tab 92 \tab 42 \tab 16 \tab 3 \tab 73 \tab 34 \tab 2 \tab 3 \tab - \tab 111 \tab 20 \tab 44 \tab 13 \tab 76 \tab 51 \tab 5 \tab - \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab 3 \tab - \tab 12 \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 8 \tab 3 \tab 1 \tab 11 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 981 \tab 97 \tab 101 \tab 152 \tab 32 \tab 5 \tab 31 \tab 54 \tab 5 \tab 1 \tab 121 \tab 107 \tab 21 \tab 3 \tab 29 \tab 16 \tab 87 \tab 59 \tab 16 \tab 3 \tab 67 \tab 41 \tab 2 \tab 4 \tab - \tab 86 \tab 19 \tab 40 \tab 13 \tab 43 \tab 46 \tab 5 \tab - \tab 4 \tab 1 \tab 7 \tab 1 \tab - \tab 2 \tab 1 \tab 11 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 7 \tab 1 \tab - \tab 8 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1403 \tab 76 \tab 168 \tab 249 \tab 50 \tab 4 \tab 64 \tab 94 \tab 11 \tab 4 \tab 218 \tab 210 \tab 31 \tab 2 \tab 58 \tab 17 \tab 172 \tab 107 \tab 32 \tab 4 \tab 143 \tab 100 \tab 5 \tab 4 \tab 1 \tab 161 \tab 38 \tab 103 \tab 37 \tab 91 \tab 91 \tab 14 \tab 2 \tab 2 \tab 58 \tab 4 \tab 4 \tab 1 \tab 7 \tab 4 \tab 14 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 9 \tab 2 \tab 1 \tab 21 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 7 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 103 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1271 \tab 174 \tab 155 \tab 204 \tab 48 \tab 6 \tab 51 \tab 93 \tab 9 \tab 3 \tab 154 \tab 165 \tab 34 \tab 2 \tab 51 \tab 30 \tab 115 \tab 94 \tab 24 \tab 1 \tab 128 \tab 93 \tab 7 \tab 12 \tab 1 \tab 145 \tab 32 \tab 67 \tab 23 \tab 84 \tab 88 \tab 11 \tab 1 \tab 20 \tab 14 \tab 5 \tab 3 \tab - \tab 7 \tab 2 \tab 22 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 12 \tab 6 \tab 1 \tab 18 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab 3544 \cr
   Total \tab 14358 \tab 1549 \tab 1941 \tab 1551 \tab 331 \tab 33 \tab 377 \tab 661 \tab 60 \tab 19 \tab 1797 \tab 1041 \tab 164 \tab 14 \tab 311 \tab 108 \tab 1193 \tab 499 \tab 147 \tab 20 \tab 931 \tab 413 \tab 23 \tab 49 \tab 3 \tab 1305 \tab 168 \tab 568 \tab 121 \tab 667 \tab 519 \tab 67 \tab 4 \tab 80 \tab 524 \tab 27 \tab 21 \tab 3 \tab 38 \tab 14 \tab 90 \tab 7 \tab 6 \tab 4 \tab 13 \tab 3 \tab 23 \tab 2 \tab 2 \tab 7 \tab 10 \tab 12 \tab 52 \tab 15 \tab 8 \tab 67 \tab 9 \tab 7 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 9 \tab 1 \tab 8 \tab 2 \tab 1 \tab 5 \tab 2 \tab 2 \tab 2 \tab 20 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 103 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab 0 \tab 32251 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artgn}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyinfgn}},
\code{\link{buyit12}},
\code{\link{byemail}},
\code{\link{byemprob}},
\code{\link{byphone}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{gamegn}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthmag2}},
\code{\link{hlthpapr}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{huclean}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{inperson}},
\code{\link{intaccss}},
\code{\link{invest12}},
\code{\link{investgn}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobhrw}},
\code{\link{jobinfo}},
\code{\link{jobminw}},
\code{\link{jobonwww}},
\code{\link{jobpaper}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{letters}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{meetings}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{newjobgn}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numcntct}},
\code{\link{numhome}},
\code{\link{numprobs}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{peoplegn}},
\code{\link{perfrmgn}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polcamp}},
\code{\link{polcangn}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldisgn}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polfrel}},
\code{\link{polinf12}},
\code{\link{polinfgn}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{polpapr}},
\code{\link{poltough}},
\code{\link{poltv}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{travelgn}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}

Other Computers And Internet: 
\code{\link{abort12}},
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advsrch}},
\code{\link{advstu}},
\code{\link{advsup}},
\code{\link{advtchr}},
\code{\link{art30}},
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{artview}},
\code{\link{askelse}},
\code{\link{askother}},
\code{\link{askwksch}},
\code{\link{askwork}},
\code{\link{autonojb}},
\code{\link{bookmark}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{browser3}},
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{callcomp}},
\code{\link{catdrcty}},
\code{\link{chat12}},
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{cmpvirus}},
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{comphome}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{compuse}},
\code{\link{compwork}},
\code{\link{contact}},
\code{\link{cook30}},
\code{\link{cussup}},
\code{\link{doonown}},
\code{\link{doself}},
\code{\link{download}},
\code{\link{drctlink}},
\code{\link{ecomhrh}},
\code{\link{ecomhro}},
\code{\link{ecomhrw}},
\code{\link{ecomminh}},
\code{\link{ecommino}},
\code{\link{ecomminw}},
\code{\link{econ12}},
\code{\link{educ30}},
\code{\link{emailhr}},
\code{\link{emaillnk}},
\code{\link{emailmin}},
\code{\link{emailyr}},
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emgroups}},
\code{\link{emhrh}},
\code{\link{emhro}},
\code{\link{emhrw}},
\code{\link{emminh}},
\code{\link{emmino}},
\code{\link{emminw}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}},
\code{\link{enviro12}},
\code{\link{ezines}},
\code{\link{fe12}},
\code{\link{fin30}},
\code{\link{foraff12}},
\code{\link{game12}},
\code{\link{games30}},
\code{\link{govt30}},
\code{\link{gun12}},
\code{\link{health12}},
\code{\link{health30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{homeband}},
\code{\link{humor30}},
\code{\link{hyperlnk}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{intaccss}},
\code{\link{internet}},
\code{\link{intrhome}},
\code{\link{invest12}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{locate12}},
\code{\link{manbook}},
\code{\link{manual}},
\code{\link{meeting}},
\code{\link{moral12}},
\code{\link{mostloc}},
\code{\link{mp3}},
\code{\link{music30}},
\code{\link{musicbuy}},
\code{\link{musicdif}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{musiclke}},
\code{\link{musiclst}},
\code{\link{musicnew}},
\code{\link{newjob12}},
\code{\link{news12}},
\code{\link{news30}},
\code{\link{newsgrps}},
\code{\link{numhome}},
\code{\link{online}},
\code{\link{payelse}},
\code{\link{payother}},
\code{\link{paywww}},
\code{\link{people12}},
\code{\link{perhrh}},
\code{\link{perhro}},
\code{\link{perhrw}},
\code{\link{perminh}},
\code{\link{permino}},
\code{\link{perminw}},
\code{\link{person30}},
\code{\link{petition}},
\code{\link{pol30}},
\code{\link{polagree}},
\code{\link{polalter}},
\code{\link{polcam12}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polinf12}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{poltough}},
\code{\link{polwww}},
\code{\link{porn30}},
\code{\link{prefsets}},
\code{\link{processr}},
\code{\link{racrel12}},
\code{\link{relig30}},
\code{\link{schl30}},
\code{\link{sci30}},
\code{\link{sports30}},
\code{\link{srcheng}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{taxes12}},
\code{\link{tecsup}},
\code{\link{travel12}},
\code{\link{travel30}},
\code{\link{tvmov30}},
\code{\link{typeurl}},
\code{\link{upload}},
\code{\link{usehome}},
\code{\link{usewww}},
\code{\link{votefor}},
\code{\link{webable}},
\code{\link{webmob}},
\code{\link{webtv}},
\code{\link{webyr}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work12}},
\code{\link{work30}},
\code{\link{wrkhrh}},
\code{\link{wrkhro}},
\code{\link{wrkminh}},
\code{\link{wrkmino}},
\code{\link{wwwart}},
\code{\link{wwwbiz}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwcowrk}},
\code{\link{wwwfrnd}},
\code{\link{wwwgfbf}},
\code{\link{wwwgthic}},
\code{\link{wwwhome}},
\code{\link{wwwhrh}},
\code{\link{wwwhro}},
\code{\link{wwwhrw}},
\code{\link{wwwjazz}},
\code{\link{wwwline1}},
\code{\link{wwwline2}},
\code{\link{wwwlit}},
\code{\link{wwwmin}},
\code{\link{wwwminh}},
\code{\link{wwwmino}},
\code{\link{wwwminw}},
\code{\link{wwwmusic}},
\code{\link{wwwnei}},
\code{\link{wwwoldie}},
\code{\link{wwwpersn}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}},
\code{\link{wwwsp}},
\code{\link{wwwvol}},
\code{\link{wwwworld}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
