% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{wwwhrh}
\alias{wwwhrh}
\title{Hours per week use computer www}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwhrh
}
\details{
Question About how many minutes or hours a week do you use your home computer for each of the following:
E. Using the world Wide web (other than for e-mail)?
2. Hours per week?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2565/vshow}{GSS Data Explorer page for \code{wwwhrh}}.
}

\seealso{
Other Information Society: 
\code{\link{compwork}},
\code{\link{emhrh}},
\code{\link{emminh}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{perhrh}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
