% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_284.R
\name{wwwhro}
\alias{wwwhro}
\title{Hours per week using other computer www}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwhro
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
E. Using the World Wide Web other than for e-mail?
2. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2594/vshow}{GSS Data Explorer page for \code{wwwhro}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab no answer \tab 10 \tab 15 \tab 20 \tab 22 \tab 37 \tab 8 \tab don't know \tab Total \cr
   2000 \tab 2665 \tab 57 \tab 22 \tab 8 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 53 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2270 \tab 251 \tab 80 \tab 51 \tab 19 \tab - \tab 10 \tab 12 \tab 3 \tab 2 \tab 44 \tab 10 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 2765 \cr
   Total \tab 4935 \tab 308 \tab 102 \tab 59 \tab 23 \tab 1 \tab 12 \tab 14 \tab 5 \tab 3 \tab 97 \tab 10 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emhro}},
\code{\link{emmino}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{wrkhro}},
\code{\link{wrkmino}},
\code{\link{wwwmino}}
}
\concept{Information Society}
\keyword{variable}
