% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{wwwhrw}
\alias{wwwhrw}
\title{Hours per week using work computer www}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwhrw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
F. Using the World Wide Web (other than for e-mail)?
2. Hours per week?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2578/vshow}{GSS Data Explorer page for \code{wwwhrw}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 15 \tab 2 \tab 20 \tab 22 \tab 26 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab don't know \tab no answer \tab 13 \tab 16 \tab 17 \tab 18 \tab 24 \tab 25 \tab 34 \tab 35 \tab 40 \tab 45 \tab 50 \tab 60 \tab 7 \tab 8 \tab 9 \tab Total \cr
   2000 \tab 2486 \tab 149 \tab 28 \tab 11 \tab 1 \tab 1 \tab 24 \tab 2 \tab 1 \tab 1 \tab 12 \tab 3 \tab 6 \tab 10 \tab 3 \tab 1 \tab 78 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1572 \tab 579 \tab 155 \tab 32 \tab 3 \tab 16 \tab 91 \tab 14 \tab 1 \tab - \tab 59 \tab 6 \tab 36 \tab 68 \tab 16 \tab 8 \tab 33 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 2 \tab 6 \tab 3 \tab 3 \tab 1 \tab 17 \tab 19 \tab 11 \tab 2765 \cr
   Total \tab 4058 \tab 728 \tab 183 \tab 43 \tab 4 \tab 17 \tab 115 \tab 16 \tab 2 \tab 1 \tab 71 \tab 9 \tab 42 \tab 78 \tab 19 \tab 9 \tab 111 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 2 \tab 6 \tab 3 \tab 3 \tab 1 \tab 17 \tab 19 \tab 11 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{ecomminw}},
\code{\link{emhrw}},
\code{\link{emminw}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{perhrw}},
\code{\link{perminw}},
\code{\link{wwwminw}}
}
\concept{Information Society}
\keyword{variable}
