% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{wwwline1}
\alias{wwwline1}
\title{Is www through phone line}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwline1
}
\details{
Question C. At your home are you connected to the World Wide Web through a regular telephone line connected to your computer, or through some other means?
}
\section{Values}{

\itemize{
\item \verb{[1]} regular telephone line
\item \verb{[2]} other means
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2554/vshow}{GSS Data Explorer page for \code{wwwline1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab regular telephone line \tab other means \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2471 \tab 302 \tab 20 \tab 1 \tab 23 \tab 2817 \cr
   2002 \tab 2098 \tab 536 \tab 119 \tab 1 \tab 11 \tab 2765 \cr
   Total \tab 4569 \tab 838 \tab 139 \tab 2 \tab 34 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
