% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{wwwline2}
\alias{wwwline2}
\title{Means to www}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwline2
}
\details{
Question D. What is this other means by which you are connected to the World Wide Web?
}
\section{Values}{

\itemize{
\item \verb{[1]} special high-speed tele line (isdn) cncted to cmptr
\item \verb{[2]} cable service line cnctd to computer
\item \verb{[3]} web tv line, cnctd to tv set
\item \verb{[4]} more than one type of cnction
\item \verb{[5]} other, not named above
\item \verb{[6]} wireless connection (e.g. from a satellite)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2555/vshow}{GSS Data Explorer page for \code{wwwline2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab special high-speed tele line (isdn) cncted to cmptr \tab cable service line cnctd to computer \tab web tv line, cnctd to tv set \tab more than one type of cnction \tab other, not named above \tab don't know \tab no answer \tab wireless connection (e.g. from a satellite) \tab Total \cr
   2000 \tab 2773 \tab 6 \tab 10 \tab 1 \tab 1 \tab 2 \tab 1 \tab 23 \tab - \tab 2817 \cr
   2002 \tab 2635 \tab 38 \tab 72 \tab 3 \tab 1 \tab 1 \tab - \tab 11 \tab 4 \tab 2765 \cr
   Total \tab 5408 \tab 44 \tab 82 \tab 4 \tab 2 \tab 3 \tab 1 \tab 34 \tab 4 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}}
}
\concept{Information Society}
\keyword{variable}
