% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_284.R
\name{wwwmino}
\alias{wwwmino}
\title{Minutes per week using other computer www}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwmino
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
E. Using the World Wide Web other than for e-mail?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2593/vshow}{GSS Data Explorer page for \code{wwwmino}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 15 \tab 2 \tab 20 \tab 30 \tab 5 \tab no answer \tab 25 \tab 3 \tab 40 \tab 45 \tab 6 \tab don't know \tab Total \cr
   2000 \tab 2665 \tab 90 \tab 1 \tab 3 \tab 3 \tab 1 \tab 3 \tab 12 \tab 5 \tab 34 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2270 \tab 326 \tab 7 \tab 13 \tab 14 \tab 3 \tab 9 \tab 54 \tab 14 \tab 44 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2765 \cr
   Total \tab 4935 \tab 416 \tab 8 \tab 16 \tab 17 \tab 4 \tab 12 \tab 66 \tab 19 \tab 78 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emhro}},
\code{\link{emmino}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{wrkhro}},
\code{\link{wrkmino}},
\code{\link{wwwhro}}
}
\concept{Information Society}
\keyword{variable}
