% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{wwwminw}
\alias{wwwminw}
\title{Minutes per week using work computer www}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwminw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
F. Using the World Wide Web (other than for e-mail)?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2577/vshow}{GSS Data Explorer page for \code{wwwminw}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 15 \tab 20 \tab 30 \tab 35 \tab 45 \tab 5 \tab don't know \tab no answer \tab 1 \tab 19 \tab 25 \tab 3 \tab 40 \tab 50 \tab 6 \tab Total \cr
   2000 \tab 2486 \tab 244 \tab 3 \tab 4 \tab 2 \tab 20 \tab 1 \tab 1 \tab 2 \tab 1 \tab 53 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1572 \tab 1022 \tab 13 \tab 17 \tab 11 \tab 63 \tab - \tab 4 \tab 8 \tab 8 \tab 33 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2765 \cr
   Total \tab 4058 \tab 1266 \tab 16 \tab 21 \tab 13 \tab 83 \tab 1 \tab 5 \tab 10 \tab 9 \tab 86 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{ecomminw}},
\code{\link{emhrw}},
\code{\link{emminw}},
\code{\link{inchrw}},
\code{\link{incminw}},
\code{\link{perhrw}},
\code{\link{perminw}},
\code{\link{wwwhrw}}
}
\concept{Information Society}
\keyword{variable}
