% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{wwwmusic}
\alias{wwwmusic}
\title{Use the web to learn about music}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwmusic
}
\details{
Question A. Do you ever use the Web to learn about, listen to, or discuss topics related to any kind of music?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2456/vshow}{GSS Data Explorer page for \code{wwwmusic}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2326 \tab 206 \tab 172 \tab 113 \tab 2817 \cr
   2002 \tab 2284 \tab 227 \tab 208 \tab 46 \tab 2765 \cr
   Total \tab 4610 \tab 433 \tab 380 \tab 159 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}}
}
\concept{Information Society}
\keyword{variable}
