% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{wwwworld}
\alias{wwwworld}
\title{In last month listen world music on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwworld
}
\details{
Question In the last month, how many times have you used the Web to learn about or listen to the following kinds of music?
I. World Music?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} 1-2 times
\item \verb{[3]} more than 2 times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2469/vshow}{GSS Data Explorer page for \code{wwwworld}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab never \tab 1-2 times \tab more than 2 times \tab no answer \tab Total \cr
   2000 \tab 2498 \tab 136 \tab 45 \tab 25 \tab 113 \tab 2817 \cr
   Total \tab 2498 \tab 136 \tab 45 \tab 25 \tab 113 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{musicnew}},
\code{\link{wwwart}}
}
\concept{Information Society}
\keyword{variable}
