% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{xhaustn}
\alias{xhaustn}
\title{How often does work exhaust r}
\source{
General Social Survey https://gss.norc.org
}
\description{
xhaustn
}
\details{
Question Now some more questions about your working conditions. Please circle one code for each item below to show how often it applies to your work. How often...
A. Do you come home from work exhausted?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4550/vshow}{GSS Data Explorer page for \code{xhaustn}}.
}

\seealso{
Other Work Orientation: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
