% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{xmovie1}
\alias{xmovie1}
\title{Seen porn film (theater or vcr) last yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
xmovie1
}
\details{
Question B. Have you seen an X-rated movie either in a movie theater or on a VCR (Video Cassette Recorder) in the last year?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/642/vshow}{GSS Data Explorer page for \code{xmovie1}}.
}

\seealso{
Other Core: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}}

Other Movies: 
\code{\link{xmovie}}

Other Sex: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{xmovie}}

Other Split Ballots: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}}
}
\concept{Core}
\concept{Movies}
\concept{Sex}
\concept{Split Ballots}
\keyword{variable}
