% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{yearly}
\alias{yearly}
\title{What is your yearly wage before tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
yearly
}
\details{
Question IF HOURLY WAGE:
E. What is your 1998 annual salary, before any taxes or deductions?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2165/vshow}{GSS Data Explorer page for \code{yearly}}.
}

\seealso{
Other Jobs: 
\code{\link{daily}},
\code{\link{hourly}},
\code{\link{monthly}},
\code{\link{trusting}},
\code{\link{weekly}}

Other Job: 
\code{\link{daily}},
\code{\link{hourly}},
\code{\link{monthly}},
\code{\link{weekly}}

Other Occupation: 
\code{\link{daily}},
\code{\link{hourly}},
\code{\link{monthly}},
\code{\link{weekly}}
}
\concept{Job}
\concept{Jobs}
\concept{Occupation}
\keyword{variable}
