% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{yearsusa}
\alias{yearsusa}
\title{Lived in u.s., years:}
\source{
General Social Survey https://gss.norc.org
}
\description{
yearsusa
}
\details{
Question How long have you lived in the United States? If you have lived in the United States for more than one period during your life, what is the total number of years you have lived in the United States? - Years
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7297/vshow}{GSS Data Explorer page for \code{yearsusa}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 6 \tab 60 \tab 61 \tab 62 \tab 68 \tab 69 \tab 7 \tab 70 \tab 72 \tab 78 \tab 8 \tab 9 \tab 93 \tab no answer \tab 41 \tab 42 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 71 \tab 74 \tab 83 \tab don't know \tab skipped on web \tab Total \cr
   2018 \tab 2049 \tab 4 \tab 10 \tab 4 \tab 11 \tab 5 \tab 7 \tab 14 \tab 8 \tab 6 \tab 14 \tab 10 \tab 7 \tab 15 \tab 2 \tab 8 \tab 3 \tab 8 \tab 7 \tab 5 \tab 8 \tab 3 \tab 6 \tab 7 \tab 6 \tab 1 \tab 3 \tab 3 \tab 4 \tab 5 \tab 3 \tab 4 \tab 6 \tab 2 \tab 4 \tab 6 \tab 4 \tab 2 \tab 5 \tab 2 \tab 6 \tab 5 \tab 5 \tab 8 \tab 6 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3567 \tab 4 \tab 11 \tab 4 \tab 12 \tab 9 \tab 8 \tab 12 \tab 10 \tab - \tab 14 \tab 11 \tab 10 \tab 22 \tab 6 \tab 9 \tab 6 \tab 3 \tab 12 \tab 7 \tab 3 \tab 6 \tab 9 \tab 14 \tab 17 \tab 6 \tab 8 \tab 8 \tab 7 \tab 7 \tab 2 \tab 4 \tab 2 \tab 5 \tab 14 \tab 13 \tab 3 \tab 8 \tab 2 \tab 1 \tab 5 \tab 2 \tab 2 \tab 17 \tab 9 \tab 2 \tab 2 \tab 5 \tab 7 \tab 3 \tab 4 \tab - \tab 1 \tab 9 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 11 \tab 6 \tab - \tab - \tab 9 \tab 11 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 21 \tab 4032 \cr
   2022 \tab 3188 \tab 8 \tab 12 \tab 2 \tab 5 \tab 7 \tab 4 \tab 8 \tab 12 \tab 8 \tab 7 \tab 7 \tab 6 \tab 14 \tab 4 \tab 9 \tab 9 \tab 9 \tab 15 \tab 5 \tab 4 \tab 3 \tab 6 \tab 7 \tab 9 \tab 5 \tab 4 \tab 6 \tab 4 \tab 5 \tab 3 \tab 5 \tab 2 \tab 2 \tab 10 \tab 8 \tab 2 \tab 4 \tab 7 \tab 6 \tab 3 \tab 2 \tab - \tab 13 \tab 8 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 9 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 7 \tab - \tab - \tab - \tab 14 \tab 6 \tab - \tab 6 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab 3544 \cr
   Total \tab 8804 \tab 16 \tab 33 \tab 10 \tab 28 \tab 21 \tab 19 \tab 34 \tab 30 \tab 14 \tab 35 \tab 28 \tab 23 \tab 51 \tab 12 \tab 26 \tab 18 \tab 20 \tab 34 \tab 17 \tab 15 \tab 12 \tab 21 \tab 28 \tab 32 \tab 12 \tab 15 \tab 17 \tab 15 \tab 17 \tab 8 \tab 13 \tab 10 \tab 9 \tab 28 \tab 27 \tab 9 \tab 14 \tab 14 \tab 9 \tab 14 \tab 9 \tab 7 \tab 38 \tab 23 \tab 5 \tab 6 \tab 9 \tab 10 \tab 5 \tab 8 \tab 2 \tab 3 \tab 23 \tab 2 \tab 4 \tab 6 \tab 2 \tab 2 \tab 18 \tab 2 \tab 2 \tab 1 \tab 27 \tab 14 \tab 1 \tab 8 \tab 11 \tab 16 \tab 4 \tab 1 \tab 5 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab 25 \tab 9924 \cr
}
}

\keyword{variable}
