% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{yearval1}
\alias{yearval1}
\title{Estimate of dollar value of payments in that year}
\source{
General Social Survey https://gss.norc.org
}
\description{
yearval1
}
\details{
Question What was the approximate total dollar value of these various payments in that year? NOTE: All responses who give valid answers to EXTRAYR are asked YEARVAL.
}
\section{Values}{

\itemize{
\item \verb{[1]} less than $100
\item \verb{[2]} $100 to $299
\item \verb{[3]} $300 to $499
\item \verb{[4]} $500 to $999
\item \verb{[5]} $1,000 to $1,499
\item \verb{[6]} $1,500 to $2,499
\item \verb{[7]} $2,500 to $3,999
\item \verb{[8]} $4,000 to $6,999
\item \verb{[9]} $7,000 to $14,999
\item \verb{[10]} $15,000 to $49,999
\item \verb{[11]} $50,000 to $99,999
\item \verb{[12]} over $100,000
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2014, 2018 \cr
   B/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8039/vshow}{GSS Data Explorer page for \code{yearval1}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab not available in this year \tab less than $100 \tab $100 to $299 \tab $300 to $499 \tab $500 to $999 \tab $1,000 to $1,499 \tab $1,500 to $2,499 \tab $2,500 to $3,999 \tab $4,000 to $6,999 \tab $7,000 to $14,999 \tab $15,000 to $49,999 \tab $50,000 to $99,999 \tab over $100,000 \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab - \tab 9 \tab 11 \tab 10 \tab 15 \tab 8 \tab 11 \tab 8 \tab 6 \tab 6 \tab 7 \tab 1 \tab 6 \tab 12 \tab 2633 \tab 22 \tab - \tab 2765 \cr
   2006 \tab - \tab 21 \tab 15 \tab 10 \tab 11 \tab 12 \tab 7 \tab 8 \tab 5 \tab 5 \tab 3 \tab 1 \tab 1 \tab 46 \tab 4334 \tab 31 \tab - \tab 4510 \cr
   2014 \tab - \tab 17 \tab 19 \tab 13 \tab 33 \tab 16 \tab 26 \tab 21 \tab 27 \tab 19 \tab 13 \tab 2 \tab 5 \tab - \tab 2325 \tab 2 \tab - \tab 2538 \cr
   2018 \tab - \tab 8 \tab 14 \tab 6 \tab 7 \tab 13 \tab 5 \tab 8 \tab 5 \tab 6 \tab 1 \tab - \tab - \tab 12 \tab 2248 \tab 15 \tab - \tab 2348 \cr
   2022 \tab - \tab 13 \tab 9 \tab 4 \tab 15 \tab 11 \tab 4 \tab 4 \tab 6 \tab 3 \tab 3 \tab - \tab 2 \tab 34 \tab 3429 \tab 3 \tab 4 \tab 3544 \cr
   Total \tab 0 \tab 68 \tab 68 \tab 43 \tab 81 \tab 60 \tab 53 \tab 49 \tab 49 \tab 39 \tab 27 \tab 4 \tab 14 \tab 104 \tab 14969 \tab 73 \tab 4 \tab 15705 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{cowrkhrd}},
\code{\link{deptperf}},
\code{\link{donothng}},
\code{\link{extr2001}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{ownstock}},
\code{\link{seecowrk}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{workfor}},
\code{\link{workfor1}},
\code{\link{yearval}}
}
\concept{Employee Compensation}
\keyword{variable}
