% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_142.R
\name{youmoney}
\alias{youmoney}
\title{Largest decision in which you had the final say}
\source{
General Social Survey https://gss.norc.org
}
\description{
youmoney
}
\details{
Question About how much money was involved in the largest decision in which you had the final say?
}
\section{Values}{

\itemize{
\item \code{9999996} $9999996 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1334/vshow}{GSS Data Explorer page for \code{youmoney}}.
}

\seealso{
Other Work Organizations: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{totmoney}},
\code{\link{wrkwell}}

Other Job: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{totmoney}}

Other Work: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{totmoney}},
\code{\link{wrkwell}}
}
\concept{Job}
\concept{Work}
\concept{Work Organizations}
\keyword{variable}
