% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{youngen}
\alias{youngen}
\title{What should younger generation be taught}
\source{
General Social Survey https://gss.norc.org
}
\description{
youngen
}
\details{
Question Here are two statements. Will you tell me which one you agree with more?
A. First, the younger generation should be taught by their elders to do what is right.
B. Second, the younger generation should be taught to think for themselves even though they may do something their elders disapprove of.
}
\section{Values}{

\itemize{
\item \code{1} what elders think
\item \code{2} think for selves
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1980 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/520/vshow}{GSS Data Explorer page for \code{youngen}}.
}

\seealso{
Other Core: 
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{helpoth}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socrel}},
\code{\link{thnkself}},
\code{\link{workhard}}
}
\concept{Core}
\keyword{variable}
