% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_182.R
\name{yourself}
\alias{yourself}
\title{R's life improved}
\source{
General Social Survey https://gss.norc.org
}
\description{
yourself
}
\details{
Question We'd like your views on how the women's movement has affected certain groups. For each group I name, please tell me whether you think the women's movement has improved their lives, made their lives worse, or had no effect On their lives.
F. You, yourself
}
\section{Values}{

\itemize{
\item \verb{[1]} improved my life
\item \verb{[2]} made my life worse
\item \verb{[3]} had no effect on my life
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1705/vshow}{GSS Data Explorer page for \code{yourself}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab improved my life \tab made my life worse \tab had no effect on my life \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 546 \tab 71 \tab 790 \tab 50 \tab 3 \tab 2904 \cr
   Total \tab 1444 \tab 546 \tab 71 \tab 790 \tab 50 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{children}},
\code{\link{manprof}},
\code{\link{men}}
}
\concept{Gender}
\keyword{variable}
