% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{yousup}
\alias{yousup}
\title{How many people do you supervise directly}
\source{
General Social Survey https://gss.norc.org
}
\description{
yousup
}
\details{
Question As an official part of your main job, do you supervise the work of other employees or tell other employees what work to do?
A. How many people Do you supervise directly?
}
\section{Values}{

\itemize{
\item \code{0} 0 people
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1314/vshow}{GSS Data Explorer page for \code{yousup}}.
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{totsup}}
}
\concept{Work Organizations}
\keyword{variable}
