% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_655.R
\name{yrarmus}
\alias{yrarmus}
\title{Archived livestreamed music or theater in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrarmus
}
\details{
Question (During the last 12 months, did you...) watch or listen to an archived web event featuring a music, theater, or dance performance? This may have included video or audio clips on, for example, YouTube, an organization's website, or a social media platform.

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7874/vshow}{GSS Data Explorer page for \code{yrarmus}}.
}

\seealso{
Other NEA: 
\code{\link{cvdart}},
\code{\link{cvdcreat}},
\code{\link{cvdlvmus}},
\code{\link{cvdmov}},
\code{\link{cvdrdg}},
\code{\link{yrarpo}},
\code{\link{yrclass}},
\code{\link{yrpod}},
\code{\link{yrstpo}}
}
\concept{NEA}
\keyword{variable}
