% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{yrartxbt}
\alias{yrartxbt}
\title{Attend art exhibit in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrartxbt
}
\details{
Question (During the last 12 months, did you...) go in person to an art exhibit, such as paintings, sculpture, textiles, graphic design, or photography?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7867/vshow}{GSS Data Explorer page for \code{yrartxbt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2022 \tab 2701 \tab 277 \tab 559 \tab 3 \tab 4 \tab 3544 \cr
   Total \tab 2701 \tab 277 \tab 559 \tab 3 \tab 4 \tab 3544 \cr
}
}

\seealso{
Other NEA: 
\code{\link{yrlvmus}}
}
\concept{NEA}
\keyword{variable}
