% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_356.R
\name{yrchurh2}
\alias{yrchurh2}
\title{How many years r born a member of church group2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrchurh2
}
\details{
Question O. How many years have you been a member of
2. Second church affiliated group
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3206/vshow}{GSS Data Explorer page for \code{yrchurh2}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 13 \tab 14 \tab 15 \tab 2 \tab 20 \tab 24 \tab 25 \tab 26 \tab 28 \tab 3 \tab 30 \tab 4 \tab 43 \tab 5 \tab 56 \tab 6 \tab 7 \tab no answer \tab Total \cr
   2004 \tab 2747 \tab 2 \tab 12 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 13 \tab 2812 \cr
   Total \tab 2747 \tab 2 \tab 12 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 13 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrchurh1}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrother1}},
\code{\link{yrother2}},
\code{\link{yrother3}},
\code{\link{yrprof6}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
