% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_354.R
\name{yrfarm1}
\alias{yrfarm1}
\title{How many years r born a member of farm organization1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrfarm1
}
\details{
Question L. How many years have you been a member of
\enumerate{
\item First farm organization
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3192/vshow}{GSS Data Explorer page for \code{yrfarm1}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 18 \tab 19 \tab 2 \tab 20 \tab 3 \tab 30 \tab 35 \tab 40 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   2004 \tab 2763 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2812 \cr
   Total \tab 2763 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfarm2}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrnat1}},
\code{\link{yrnat2}},
\code{\link{yrnat3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
