% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_355.R
\name{yrfarm3}
\alias{yrfarm3}
\title{How many years r born a member of farm organization3}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrfarm3
}
\details{
Question L. How many years have you been a member of
3. Third farm organization
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3194/vshow}{GSS Data Explorer page for \code{yrfarm3}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab 20 \tab no answer \tab Total \cr
   2004 \tab 2806 \tab 1 \tab 5 \tab 2812 \cr
   Total \tab 2806 \tab 1 \tab 5 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
