% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_350.R
\name{yrfrat1}
\alias{yrfrat1}
\title{How many yrs r been a member of fraternal groups1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrfrat1
}
\details{
Question A. How many years have you been a member of
\enumerate{
\item First fraternal group
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3144/vshow}{GSS Data Explorer page for \code{yrfrat1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 14 \tab 15 \tab 2 \tab 20 \tab 25 \tab 26 \tab 29 \tab 3 \tab 30 \tab 32 \tab 33 \tab 4 \tab 40 \tab 45 \tab 47 \tab 5 \tab 50 \tab 53 \tab 55 \tab 57 \tab 6 \tab 7 \tab 71 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2707 \tab 5 \tab 8 \tab 6 \tab 2 \tab 3 \tab 2 \tab 7 \tab 5 \tab 6 \tab 3 \tab 1 \tab 1 \tab 5 \tab 8 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 9 \tab 2812 \cr
   Total \tab 2707 \tab 5 \tab 8 \tab 6 \tab 2 \tab 3 \tab 2 \tab 7 \tab 5 \tab 6 \tab 3 \tab 1 \tab 1 \tab 5 \tab 8 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrserv1}},
\code{\link{yrserv2}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}},
\code{\link{yrserv6}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
