% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_354.R
\name{yrgreek1}
\alias{yrgreek1}
\title{How many years r born a member of school fraternity1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrgreek1
}
\details{
Question J. How many years have you been a member of
\enumerate{
\item First school fraternity or sorority
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3186/vshow}{GSS Data Explorer page for \code{yrgreek1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 14 \tab 15 \tab 17 \tab 2 \tab 20 \tab 21 \tab 25 \tab 28 \tab 3 \tab 30 \tab 31 \tab 32 \tab 37 \tab 4 \tab 40 \tab 5 \tab 50 \tab 53 \tab 56 \tab 64 \tab 7 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2746 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 6 \tab 1 \tab 4 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 2746 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 6 \tab 1 \tab 4 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrnat1}},
\code{\link{yrnat2}},
\code{\link{yrnat3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
