% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_357.R
\name{yrinfrm2}
\alias{yrinfrm2}
\title{How many years r born a member of informal group2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrinfrm2
}
\details{
Question Q. How many years have you been a member of…?
2. Second informal group
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3217/vshow}{GSS Data Explorer page for \code{yrinfrm2}}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 15 \tab 16 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 36 \tab 4 \tab 5 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2708 \tab 5 \tab 13 \tab 7 \tab 1 \tab 6 \tab 3 \tab 1 \tab 12 \tab 2 \tab 1 \tab 10 \tab 4 \tab 1 \tab 4 \tab 5 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 17 \tab 2812 \cr
   Total \tab 2708 \tab 5 \tab 13 \tab 7 \tab 1 \tab 6 \tab 3 \tab 1 \tab 12 \tab 2 \tab 1 \tab 10 \tab 4 \tab 1 \tab 4 \tab 5 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 17 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrother4}},
\code{\link{yrother5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
