% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_355.R
\name{yrlit4}
\alias{yrlit4}
\title{How many years r born a member of literary group4}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrlit4
}
\details{
Question M. How many years have you been a member of
4. Fourth literary, art, discussion, or study groups
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3198/vshow}{GSS Data Explorer page for \code{yrlit4}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab 2 \tab no answer \tab Total \cr
   2004 \tab 2802 \tab 1 \tab 9 \tab 2812 \cr
   Total \tab 2802 \tab 1 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfarm3}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
