% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{yrlvmus}
\alias{yrlvmus}
\title{Attend live music or theater in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrlvmus
}
\details{
Question First, we'd like to ask about some leisure or recreational activities that people do during their free time. As we list each activity, please indicate if it is something you have done in the past 12 months. During the last 12 months, did you go in person to a live music, theater, or dance performance?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7866/vshow}{GSS Data Explorer page for \code{yrlvmus}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab Total \cr
   2022 \tab 2701 \tab 431 \tab 408 \tab 4 \tab 3544 \cr
   Total \tab 2701 \tab 431 \tab 408 \tab 4 \tab 3544 \cr
}
}

\seealso{
Other NEA: 
\code{\link{yrartxbt}}
}
\concept{NEA}
\keyword{variable}
