% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_654.R
\name{yrmovie}
\alias{yrmovie}
\title{See movie in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrmovie
}
\details{
Question (During the last 12 months, did you...) go to the movies?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7868/vshow}{GSS Data Explorer page for \code{yrmovie}}.
}

\seealso{
Other NEA: 
\code{\link{artsout}},
\code{\link{cvdarmus}},
\code{\link{cvdarpo}},
\code{\link{cvdart}},
\code{\link{cvdclass}},
\code{\link{cvdcreat}},
\code{\link{cvdlvmus}},
\code{\link{cvdmov}},
\code{\link{cvdpod}},
\code{\link{cvdrdg}},
\code{\link{cvdstmus}},
\code{\link{cvdstpo}},
\code{\link{cvdtour}},
\code{\link{neaeligible}},
\code{\link{yrarmus}},
\code{\link{yrarpo}},
\code{\link{yrartxbt}},
\code{\link{yrclass}},
\code{\link{yrcreat}},
\code{\link{yrlvmus}},
\code{\link{yrpod}},
\code{\link{yrrdg}},
\code{\link{yrstmus}},
\code{\link{yrstpo}},
\code{\link{yrtour}}
}
\concept{NEA}
\keyword{variable}
