% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_354.R
\name{yrnat1}
\alias{yrnat1}
\title{How many years r born a member of national group1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrnat1
}
\details{
Question K. How many years have you been a member of…?
\enumerate{
\item First nationality group
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3189/vshow}{GSS Data Explorer page for \code{yrnat1}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 15 \tab 18 \tab 19 \tab 2 \tab 20 \tab 24 \tab 3 \tab 30 \tab 35 \tab 4 \tab 40 \tab 5 \tab 6 \tab 8 \tab no answer \tab Total \cr
   2004 \tab 2768 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 8 \tab 2812 \cr
   Total \tab 2768 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrnat2}},
\code{\link{yrnat3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
