% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_354.R
\name{yrnat2}
\alias{yrnat2}
\title{How many years r born a member of national group2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrnat2
}
\details{
Question K. How many years have you been a member of…?
2. Second nationality group
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3190/vshow}{GSS Data Explorer page for \code{yrnat2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 4 \tab 5 \tab no answer \tab Total \cr
   2004 \tab 2804 \tab 1 \tab 1 \tab 6 \tab 2812 \cr
   Total \tab 2804 \tab 1 \tab 1 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrnat1}},
\code{\link{yrnat3}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
