% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_356.R
\name{yrother1}
\alias{yrother1}
\title{How many years r born a member of other group1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrother1
}
\details{
Question P. How many years have you been a member of…?
\enumerate{
\item First other group
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3211/vshow}{GSS Data Explorer page for \code{yrother1}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 15 \tab 16 \tab 2 \tab 20 \tab 22 \tab 27 \tab 3 \tab 4 \tab 40 \tab 5 \tab 6 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2706 \tab 29 \tab 14 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 10 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 11 \tab 17 \tab 2812 \cr
   Total \tab 2706 \tab 29 \tab 14 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 10 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 11 \tab 17 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrchurh1}},
\code{\link{yrchurh2}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrother2}},
\code{\link{yrother3}},
\code{\link{yrprof6}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
