% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_356.R
\name{yrother2}
\alias{yrother2}
\title{How many years r born a member of other group2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrother2
}
\details{
Question P. How many years have you been a member of…?
2. Second other group
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3212/vshow}{GSS Data Explorer page for \code{yrother2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 15 \tab 24 \tab 7 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2794 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2812 \cr
   Total \tab 2794 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrchurh1}},
\code{\link{yrchurh2}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrother1}},
\code{\link{yrother3}},
\code{\link{yrprof6}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
