% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_655.R
\name{yrpod}
\alias{yrpod}
\title{Listen to art podcasts in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrpod
}
\details{
Question (During the last 12 months, did you...) listen to podcasts about music, theater, or dance, or about visual arts or film, or about books or literature?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7878/vshow}{GSS Data Explorer page for \code{yrpod}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2022 \tab 2701 \tab 238 \tab 600 \tab 1 \tab 4 \tab 3544 \cr
   Total \tab 2701 \tab 238 \tab 600 \tab 1 \tab 4 \tab 3544 \cr
}
}

\seealso{
Other NEA: 
\code{\link{cvdarmus}},
\code{\link{cvdart}},
\code{\link{cvdcreat}},
\code{\link{cvdlvmus}},
\code{\link{cvdmov}},
\code{\link{cvdrdg}},
\code{\link{cvdstmus}},
\code{\link{cvdstpo}},
\code{\link{cvdtour}}
}
\concept{NEA}
\keyword{variable}
