% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_351.R
\name{yrpolit1}
\alias{yrpolit1}
\title{How many years r been a member of political club1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrpolit1
}
\details{
Question D. How many years have you been a member of…?
\enumerate{
\item First political club
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3161/vshow}{GSS Data Explorer page for \code{yrpolit1}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 14 \tab 15 \tab 18 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 35 \tab 4 \tab 40 \tab 5 \tab 50 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2744 \tab 2 \tab 12 \tab 12 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 5 \tab 7 \tab 2812 \cr
   Total \tab 2744 \tab 2 \tab 12 \tab 12 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 5 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
