% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_355.R
\name{yrprof1}
\alias{yrprof1}
\title{How many years r born a member of professional society1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrprof1
}
\details{
Question N. How many years have you been a member of
\enumerate{
\item First professional or academic society
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3199/vshow}{GSS Data Explorer page for \code{yrprof1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 25 \tab 26 \tab 27 \tab 28 \tab 3 \tab 30 \tab 32 \tab 33 \tab 35 \tab 4 \tab 40 \tab 5 \tab 52 \tab 56 \tab 6 \tab 60 \tab 7 \tab 8 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2586 \tab 4 \tab 24 \tab 17 \tab 3 \tab 5 \tab 1 \tab 3 \tab 11 \tab 4 \tab 2 \tab 4 \tab 19 \tab 10 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 23 \tab 7 \tab 1 \tab 1 \tab 3 \tab 17 \tab 1 \tab 10 \tab 1 \tab 1 \tab 11 \tab 1 \tab 6 \tab 5 \tab 5 \tab 14 \tab 2812 \cr
   Total \tab 2586 \tab 4 \tab 24 \tab 17 \tab 3 \tab 5 \tab 1 \tab 3 \tab 11 \tab 4 \tab 2 \tab 4 \tab 19 \tab 10 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 23 \tab 7 \tab 1 \tab 1 \tab 3 \tab 17 \tab 1 \tab 10 \tab 1 \tab 1 \tab 11 \tab 1 \tab 6 \tab 5 \tab 5 \tab 14 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfarm3}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
