% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_355.R
\name{yrprof5}
\alias{yrprof5}
\title{How many years r born a member of professional society5}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrprof5
}
\details{
Question N. How many years have you been a member of
5. Fifth professional or academic society
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3203/vshow}{GSS Data Explorer page for \code{yrprof5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 1 \tab 18 \tab 4 \tab no answer \tab Total \cr
   2004 \tab 2797 \tab 1 \tab 1 \tab 2 \tab 11 \tab 2812 \cr
   Total \tab 2797 \tab 1 \tab 1 \tab 2 \tab 11 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfarm3}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
