% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_353.R
\name{yrschl1}
\alias{yrschl1}
\title{How many years r born a member of school service group1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrschl1
}
\details{
Question H. How many years have you been a member of
\enumerate{
\item First school service group
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3179/vshow}{GSS Data Explorer page for \code{yrschl1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 18 \tab 19 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 33 \tab 4 \tab 5 \tab 51 \tab 52 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2601 \tab 15 \tab 31 \tab 5 \tab 2 \tab 5 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 37 \tab 2 \tab 3 \tab 29 \tab 3 \tab 1 \tab 12 \tab 10 \tab 1 \tab 1 \tab 7 \tab 9 \tab 4 \tab 3 \tab 7 \tab 13 \tab 2812 \cr
   Total \tab 2601 \tab 15 \tab 31 \tab 5 \tab 2 \tab 5 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 37 \tab 2 \tab 3 \tab 29 \tab 3 \tab 1 \tab 12 \tab 10 \tab 1 \tab 1 \tab 7 \tab 9 \tab 4 \tab 3 \tab 7 \tab 13 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrhobby1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
