% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_350.R
\name{yrserv2}
\alias{yrserv2}
\title{How many years r been a member of service club2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrserv2
}
\details{
Question B. How many years have you been a member of
2. Second service club
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3149/vshow}{GSS Data Explorer page for \code{yrserv2}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 12 \tab 14 \tab 15 \tab 2 \tab 20 \tab 3 \tab 4 \tab 40 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   2004 \tab 2774 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 2812 \cr
   Total \tab 2774 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfrat1}},
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrserv1}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}},
\code{\link{yrserv6}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
