% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_350.R
\name{yrserv6}
\alias{yrserv6}
\title{How many years r been a member of service club6}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrserv6
}
\details{
Question B. How many years have you been a member of
6. Sixth service club
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3153/vshow}{GSS Data Explorer page for \code{yrserv6}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrfrat1}},
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrserv1}},
\code{\link{yrserv2}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
