% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{yrsfirm}
\alias{yrsfirm}
\title{Importance of tenure to pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrsfirm
}
\details{
Question Think of two people doing the same kind of work. What do you personally think should be important in deciding how much to pay them?
Looking at the things below, please write ''1'' next to the thing that you think should be most important. Then write ''2'' next to the thing that you think should be next most important, and ''3'' next to the thing that you think should be third most important. Leave the other boxes blank.
A. How long the employee has been with the firm
}
\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} 2nd most impt
\item \code{3} 3rd most impt
\item \code{4} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4496/vshow}{GSS Data Explorer page for \code{yrsfirm}}.
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{hwduties}},
\code{\link{paysame}},
\code{\link{sexemp}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}}

Other Employment: 
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{paysame}}

Other Occupation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{paysame}},
\code{\link{sexemp}}
}
\concept{Employment}
\concept{Occupation}
\concept{Work Orientation}
\keyword{variable}
