% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_366.R
\name{yrskwn2}
\alias{yrskwn2}
\title{Years r has known the named2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrskwn2
}
\details{
Question How long have you known (NAME)?
B. How many years have you known the second name?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3309/vshow}{GSS Data Explorer page for \code{yrskwn2}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{relneg}},
\code{\link{yrskwn1}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
