% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_366.R
\name{yrskwn5}
\alias{yrskwn5}
\title{Years r has known the named5}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrskwn5
}
\details{
Question How long have you known (NAME)?
E. How many years have you known the fifth name?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3312/vshow}{GSS Data Explorer page for \code{yrskwn5}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 4 \tab 40 \tab 41 \tab 44 \tab 45 \tab 46 \tab 47 \tab 49 \tab 5 \tab 50 \tab 54 \tab 57 \tab 58 \tab 6 \tab 60 \tab 61 \tab 66 \tab 7 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2600 \tab 1 \tab 11 \tab 8 \tab 2 \tab 4 \tab 2 \tab 4 \tab 5 \tab 3 \tab 1 \tab 5 \tab 8 \tab 5 \tab 2 \tab 2 \tab 4 \tab 1 \tab 8 \tab 3 \tab 1 \tab 2 \tab 2 \tab 6 \tab 9 \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 3 \tab 6 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 47 \tab 2812 \cr
   Total \tab 2600 \tab 1 \tab 11 \tab 8 \tab 2 \tab 4 \tab 2 \tab 4 \tab 5 \tab 3 \tab 1 \tab 5 \tab 8 \tab 5 \tab 2 \tab 2 \tab 4 \tab 1 \tab 8 \tab 3 \tab 1 \tab 2 \tab 2 \tab 6 \tab 9 \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 3 \tab 6 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 47 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{relneg}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
