% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_352.R
\name{yrsport2}
\alias{yrsport2}
\title{How many years r born a member of sports groups2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrsport2
}
\details{
Question F. How many years have you been a member of
2. Second sports group
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3169/vshow}{GSS Data Explorer page for \code{yrsport2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 15 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 4 \tab 40 \tab 5 \tab 6 \tab 7 \tab 8 \tab no answer \tab Total \cr
   2004 \tab 2740 \tab 2 \tab 9 \tab 4 \tab 3 \tab 4 \tab 13 \tab 2 \tab 1 \tab 6 \tab 1 \tab 6 \tab 2 \tab 2 \tab 5 \tab 1 \tab 2 \tab 9 \tab 2812 \cr
   Total \tab 2740 \tab 2 \tab 9 \tab 4 \tab 3 \tab 4 \tab 13 \tab 2 \tab 1 \tab 6 \tab 1 \tab 6 \tab 2 \tab 2 \tab 5 \tab 1 \tab 2 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrsport1}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrsport6}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
