% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_352.R
\name{yrsport5}
\alias{yrsport5}
\title{How many years r born a member of sports groups5}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrsport5
}
\details{
Question F. How many years have you been a member of
5. Fifth sports group
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3172/vshow}{GSS Data Explorer page for \code{yrsport5}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport6}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
