% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_352.R
\name{yrsport6}
\alias{yrsport6}
\title{How many years r born a member of sports groups6}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrsport6
}
\details{
Question F. How many years have you been a member of
6. Sixth sports group
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3173/vshow}{GSS Data Explorer page for \code{yrsport6}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab no answer \tab Total \cr
   2004 \tab 2804 \tab 8 \tab 2812 \cr
   Total \tab 2804 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
