% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_352.R
\name{yrunion1}
\alias{yrunion1}
\title{How many years r born a member of labor union1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrunion1
}
\details{
Question E. How many years have you been a member of
\enumerate{
\item First labor union
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3164/vshow}{GSS Data Explorer page for \code{yrunion1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 25 \tab 26 \tab 28 \tab 3 \tab 30 \tab 32 \tab 34 \tab 35 \tab 36 \tab 38 \tab 39 \tab 4 \tab 40 \tab 45 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   2004 \tab 2666 \tab 4 \tab 5 \tab 7 \tab 1 \tab 6 \tab 4 \tab 3 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 7 \tab 6 \tab 1 \tab 1 \tab 2 \tab 5 \tab 3 \tab 1 \tab 10 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 5 \tab 1 \tab 11 \tab 10 \tab 3 \tab 7 \tab 1 \tab 9 \tab 2812 \cr
   Total \tab 2666 \tab 4 \tab 5 \tab 7 \tab 1 \tab 6 \tab 4 \tab 3 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 7 \tab 6 \tab 1 \tab 1 \tab 2 \tab 5 \tab 3 \tab 1 \tab 10 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 5 \tab 1 \tab 11 \tab 10 \tab 3 \tab 7 \tab 1 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrsport6}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
