% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_351.R
\name{yrvet1}
\alias{yrvet1}
\title{How many years r been a member of veteran groups1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrvet1
}
\details{
Question C. How many years have you been a member of
\enumerate{
\item First veterans'' group
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3156/vshow}{GSS Data Explorer page for \code{yrvet1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 15 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 25 \tab 3 \tab 30 \tab 31 \tab 32 \tab 34 \tab 35 \tab 4 \tab 40 \tab 45 \tab 5 \tab 50 \tab 51 \tab 56 \tab 58 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2729 \tab 3 \tab 7 \tab 7 \tab 2 \tab 3 \tab 4 \tab 5 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 6 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 9 \tab 2812 \cr
   Total \tab 2729 \tab 3 \tab 7 \tab 7 \tab 2 \tab 3 \tab 4 \tab 5 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 6 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
