% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_351.R
\name{yrvet2}
\alias{yrvet2}
\title{How many years r been a member of veteran groups2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrvet2
}
\details{
Question C. How many years have you been a member of
2. Second veterans'' group
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3157/vshow}{GSS Data Explorer page for \code{yrvet2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 18 \tab 20 \tab 3 \tab 40 \tab 6 \tab no answer \tab Total \cr
   2004 \tab 2796 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 2796 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrvet1}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
