% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_351.R
\name{yrvet3}
\alias{yrvet3}
\title{How many years r been a member of veteran groups3}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrvet3
}
\details{
Question C. How many years have you been a member of
3. Third veterans'' group
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3158/vshow}{GSS Data Explorer page for \code{yrvet3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 18 \tab 3 \tab 6 \tab no answer \tab Total \cr
   2004 \tab 2802 \tab 1 \tab 1 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 2802 \tab 1 \tab 1 \tab 1 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet4}},
\code{\link{yrvet5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
