% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{zodiac}
\alias{zodiac}
\title{Respondents astrological sign}
\source{
General Social Survey https://gss.norc.org
}
\description{
zodiac
}
\details{
Question ASTROLOGICAL SIGN OF RESPONDENT
}
\section{Values}{

\itemize{
\item \code{1} aries
\item \code{2} taurus
\item \code{3} gemini
\item \code{4} cancer
\item \code{5} leo
\item \code{6} virgo
\item \code{7} libra
\item \code{8} scorpio
\item \code{9} sagittarius
\item \code{10} capricorn
\item \code{11} aquarius
\item \code{12} pisces
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5510/vshow}{GSS Data Explorer page for \code{zodiac}}.
}

\seealso{
Other Cohort and Birth: 
\code{\link{cohort}},
\code{\link{marcohrt}}
}
\concept{Astrological Sign}
\concept{Cohort and Birth}
\keyword{variable}
