% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_661.R
\name{abbelief}
\alias{abbelief}
\title{Pro-choice vs pro-life}
\source{
General Social Survey https://gss.norc.org
}
\description{
abbelief
}
\details{
Question Which of the following best represents you in terms of your belief about abortion? I am...

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} strongly pro-choice
\item \code{2} slightly pro-choice
\item \code{3} both pro-choice and pro-life
\item \code{4} neither pro-choice nor pro-life
\item \code{5} slightly pro-life
\item \code{6} strongly pro-life
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7947/vshow}{GSS Data Explorer page for \code{abbelief}}.
}

\seealso{
Other GSS Next: 
\code{\link{abgender}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16frndnum}},
\code{\link{brv16gpanum}},
\code{\link{brv16othnum}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{gestate}}
}
\concept{GSS Next}
\keyword{variable}
