% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{abcare}
\alias{abcare}
\title{How concerned is r about abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abcare
}
\details{
Question How concerned are you personally about the abortion issue?  Are you very concerned, somewhat concerned, not very concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \code{1} very concerned
\item \code{2} somewhat concerned
\item \code{3} not very concerned
\item \code{4} not concerned at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/618/vshow}{GSS Data Explorer page for \code{abcare}}.
}

\seealso{
Other Core: 
\code{\link{abfirm}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsoon}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{sexeduc}},
\code{\link{teenpill}}
}
\concept{Core}
\keyword{variable}
