% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{abchoose}
\alias{abchoose}
\title{Should a woman get an abortion for any reason}
\source{
General Social Survey https://gss.norc.org
}
\description{
abchoose
}
\details{
Question Do you agree or disagree . . .
C. A pregnant woman should be able to obtain a legal abortion for any reason whatsoever, if she chooses not to have the baby.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4414/vshow}{GSS Data Explorer page for \code{abchoose}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 257 \tab 349 \tab 126 \tab 287 \tab 313 \tab 34 \tab 81 \tab 2992 \cr
   Total \tab 1545 \tab 257 \tab 349 \tab 126 \tab 287 \tab 313 \tab 34 \tab 81 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
