% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{abchoose}
\alias{abchoose}
\title{Should a woman get an abortion for any reason}
\source{
General Social Survey https://gss.norc.org
}
\description{
abchoose
}
\details{
Question Do you agree or disagree . . .
C. A pregnant woman should be able to obtain a legal abortion for any reason whatsoever, if she chooses not to have the baby.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4414/vshow}{GSS Data Explorer page for \code{abchoose}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{Abortion}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
