% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{abdefct1}
\alias{abdefct1}
\title{Law should allow abortion for birth defects}
\source{
General Social Survey https://gss.norc.org
}
\description{
abdefct1
}
\details{
Question Do you think the law should or should not allow a pregnant woman to obtain a legal abortion. . .
A. If there is a strong chance of serious defect in the baby.
}
\section{Values}{

\itemize{
\item \code{1} definitely should allow it
\item \code{2} probably should allow it
\item \code{3} probably should not allow it
\item \code{4} definitely should not allow it
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4633/vshow}{GSS Data Explorer page for \code{abdefct1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should allow it \tab probably should allow it \tab probably should not allow it \tab definitely should not allow it \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 810 \tab 232 \tab 74 \tab 146 \tab 63 \tab 34 \tab 1517 \cr
   Total \tab 158 \tab 810 \tab 232 \tab 74 \tab 146 \tab 63 \tab 34 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}
}
\concept{ISSP Religion}
\keyword{variable}
