% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{abdefctw}
\alias{abdefctw}
\title{Wrong for woman to get abortion for birth defects}
\source{
General Social Survey https://gss.norc.org
}
\description{
abdefctw
}
\details{
Question Do you personally think it is wrong or not wrong for a woman to have an abortion. . .
A. If there is a strong chance of serious defect in the baby.
}
\section{Values}{

\itemize{
\item \verb{[1]} always wrong
\item \verb{[2]} almost always wrong
\item \verb{[3]} wrong only sometimes
\item \verb{[4]} not wrong at all
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4635/vshow}{GSS Data Explorer page for \code{abdefctw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always wrong \tab almost always wrong \tab wrong only sometimes \tab not wrong at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 279 \tab 131 \tab 163 \tab 662 \tab 90 \tab 34 \tab 1517 \cr
   1998 \tab 1548 \tab 275 \tab 111 \tab 159 \tab 621 \tab 95 \tab 23 \tab 2832 \cr
   2008 \tab 658 \tab 300 \tab 109 \tab 204 \tab 701 \tab 41 \tab 10 \tab 2023 \cr
   Total \tab 2364 \tab 854 \tab 351 \tab 526 \tab 1984 \tab 226 \tab 67 \tab 6372 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}
}
\concept{ISSP Religion}
\keyword{variable}
