% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{abdefect}
\alias{abdefect}
\title{Strong chance of serious defect}
\source{
General Social Survey https://gss.norc.org
}
\description{
abdefect
}
\details{
Question Please tell me whether or not you think it should be possible for a pregnant woman to obtain a legal abortion if. . .READ EACH STATEMENT, AND CIRCLE ONE CODE FOR EACH.
A. If there is a strong chance of serious defect in the baby?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/- \tab full \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
   1998 \tab A/C \tab partial \cr
   2000 \tab A/C \tab partial \cr
   2002 \tab A/C \tab partial \cr
   2004 \tab A/C \tab partial \cr
   2006 \tab A/C \tab partial \cr
   2008 \tab A/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
   2012 \tab A/C \tab partial \cr
   2014 \tab A/C \tab partial \cr
   2016 \tab A/C \tab partial \cr
   2018 \tab A/C \tab partial \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/601/vshow}{GSS Data Explorer page for \code{abdefect}}.
}

\seealso{
Other Core: 
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}

Other Abortion: 
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}}

Other Split Ballots: 
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{abrape}},
\code{\link{absingle}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}
}
\concept{Abortion}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
