% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{abhave1}
\alias{abhave1}
\title{Have abortion if probable defect}
\source{
General Social Survey https://gss.norc.org
}
\description{
abhave1
}
\details{
Question (WOMEN): Would you yourself have an abortion if…
(MEN):  Would you yourself advise the woman you were married to  have an abortion if…
A. There is a strong chance of serious defect in the baby?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/699/vshow}{GSS Data Explorer page for \code{abhave1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 1081 \tab 351 \tab 87 \tab 11 \tab 1530 \cr
   Total \tab 0 \tab 1081 \tab 351 \tab 87 \tab 11 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave2}},
\code{\link{abspno}},
\code{\link{comprend}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{form}},
\code{\link{phone}},
\code{\link{radiohrs}}
}
\concept{Core}
\keyword{variable}
