% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{abhave3}
\alias{abhave3}
\title{Have abortion if couldn't afford child}
\source{
General Social Survey https://gss.norc.org
}
\description{
abhave3
}
\details{
Question (WOMEN): Would you yourself have an abortion if…
(MEN):  Would you yourself advise the woman you were married to  have an abortion if…
C. You did not think you could afford another child?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/701/vshow}{GSS Data Explorer page for \code{abhave3}}.
}

\seealso{
Other Core: 
\code{\link{ablegal}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
