% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{abhave3}
\alias{abhave3}
\title{Have abortion if couldn't afford child}
\source{
General Social Survey https://gss.norc.org
}
\description{
abhave3
}
\details{
Question (WOMEN): Would you yourself have an abortion if…
(MEN):  Would you yourself advise the woman you were married to  have an abortion if…
C. You did not think you could afford another child?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/701/vshow}{GSS Data Explorer page for \code{abhave3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 427 \tab 1043 \tab 47 \tab 13 \tab 1530 \cr
   Total \tab 0 \tab 427 \tab 1043 \tab 47 \tab 13 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{ablegal}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
