% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_68.R
\name{abimp}
\alias{abimp}
\title{Importance of abortion issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
abimp
}
\details{
Question A. How important is the abortion issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \code{1} one of most imp
\item \code{2} important
\item \code{3} not very imp
\item \code{4} not imp at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/615/vshow}{GSS Data Explorer page for \code{abimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not imp at all \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 218 \tab 765 \tab 557 \tab 285 \tab 25 \tab 10 \tab 1860 \cr
   1984 \tab 484 \tab 160 \tab 459 \tab 264 \tab 97 \tab 7 \tab 2 \tab 1473 \cr
   Total \tab 484 \tab 378 \tab 1224 \tab 821 \tab 382 \tab 32 \tab 12 \tab 3333 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abinfo}},
\code{\link{aborct}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}}
}
\concept{Core}
\keyword{variable}
