% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_316.R
\name{abinspay}
\alias{abinspay}
\title{Should people be able to use health insurance for abortion - ver y}
\source{
General Social Survey https://gss.norc.org
}
\description{
abinspay
}
\details{
Question People use their health insurance to help cover the cost of receiving health care. Do you think people should be able to use their health insurance to help cover the cost of receiving an abortion?
}
\section{Values}{

\itemize{
\item \verb{[1]} people should be able to
\item \verb{[2]} people should not be able to
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7183/vshow}{GSS Data Explorer page for \code{abinspay}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab people should be able to \tab people should not be able to \tab don't know \tab no answer \tab Total \cr
   2018 \tab - \tab 1077 \tab 1068 \tab 137 \tab 66 \tab 2348 \cr
   Total \tab 0 \tab 1077 \tab 1068 \tab 137 \tab 66 \tab 2348 \cr
}
}

\keyword{variable}
