% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{ablegal}
\alias{ablegal}
\title{Should abortion ever be legal}
\source{
General Social Survey https://gss.norc.org
}
\description{
ablegal
}
\details{
Question Do you think abortions should be legal under any circumstances, legal only under certain circumstances, or never legal under any circumstance?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} sometimes
\item \verb{[3]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/702/vshow}{GSS Data Explorer page for \code{ablegal}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab always \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 331 \tab 1031 \tab 135 \tab 22 \tab 11 \tab 1530 \cr
   Total \tab 0 \tab 331 \tab 1031 \tab 135 \tab 22 \tab 11 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
