% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{ablegal}
\alias{ablegal}
\title{Should abortion ever be legal}
\source{
General Social Survey https://gss.norc.org
}
\description{
ablegal
}
\details{
Question Do you think abortions should be legal under any circumstances, legal only under certain circumstances, or never legal under any circumstance?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} sometimes
\item \code{3} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/702/vshow}{GSS Data Explorer page for \code{ablegal}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab always \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 331 \tab 1031 \tab 135 \tab 22 \tab 11 \tab 1530 \cr
   Total \tab 0 \tab 331 \tab 1031 \tab 135 \tab 22 \tab 11 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
