% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_316.R
\name{abmedgov1}
\alias{abmedgov1}
\title{Woman and doctor or govt should decide what info needed for abortion - ver x}
\source{
General Social Survey https://gss.norc.org
}
\description{
abmedgov1
}
\details{
Question Which of these statements comes closer to your view about what information a woman should have when she makes a decision about whether to have an abortion?
}
\section{Values}{

\itemize{
\item \code{1} a woman and her medical professional should decide
\item \code{2} the government should decide
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7181/vshow}{GSS Data Explorer page for \code{abmedgov1}}.
}

\keyword{variable}
