% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{abmoral}
\alias{abmoral}
\title{R has a moral opposition to abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
abmoral
}
\details{
Question Leaving aside whether or not you think abortion should be legal, are you morally opposed to abortion or not, or would you say it depends?
}
\section{Values}{

\itemize{
\item \code{1} morally opposed
\item \code{2} not morally opposed
\item \code{3} it depends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7174/vshow}{GSS Data Explorer page for \code{abmoral}}.
}

\seealso{
Other GSS Next: 
\code{\link{abhelp1}}
}
\concept{GSS Next}
\keyword{variable}
