% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_68.R
\name{aborct}
\alias{aborct}
\title{Heard supreme court decision on abortion}
\source{
General Social Survey https://gss.norc.org
}
\description{
aborct
}
\details{
Question Have you heard or read of the recent Supreme Court decision concerning abortion?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/608/vshow}{GSS Data Explorer page for \code{aborct}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 1288 \tab 204 \tab 11 \tab 1 \tab 1504 \cr
   Total \tab 0 \tab 1288 \tab 204 \tab 11 \tab 1 \tab 1504 \cr
}
}

\seealso{
Other Core: 
\code{\link{abany}},
\code{\link{abcon1}},
\code{\link{abcon2}},
\code{\link{abcon3}},
\code{\link{abimp}},
\code{\link{abinfo}},
\code{\link{abpro1}},
\code{\link{abpro2}},
\code{\link{abpro3}}
}
\concept{Core}
\keyword{variable}
