% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{abpoor1}
\alias{abpoor1}
\title{Law should allow abortion if family low income}
\source{
General Social Survey https://gss.norc.org
}
\description{
abpoor1
}
\details{
Question Do you think the law should or should not allow a pregnant woman to obtain a legal abortion. . .
B. If the family has a very low income and cannot afford any more children.
}
\section{Values}{

\itemize{
\item \code{1} definitely should allow it
\item \code{2} probably should allow it
\item \code{3} probably should not allow it
\item \code{4} definitely should not allow it
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4634/vshow}{GSS Data Explorer page for \code{abpoor1}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}

Other Abortion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoorw}}
}
\concept{Abortion}
\concept{ISSP Religion}
\keyword{variable}
