% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_626.R
\name{abpoorg}
\alias{abpoorg}
\title{Low income--cant afford more children (grid on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
abpoorg
}
\details{
Question (Pleae indicate whether or not you think it should be possible for a pregnant woman to obtain a legal abortion . . .) if the family has a very low income and cannot afford any more children?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7469/vshow}{GSS Data Explorer page for \code{abpoorg}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab don't know \tab no answer \tab Total \cr
   2021 \tab 2822 \tab 653 \tab 507 \tab 50 \tab - \tab - \tab 4032 \cr
   2022 \tab 2380 \tab 696 \tab 417 \tab 11 \tab 34 \tab 6 \tab 3544 \cr
   Total \tab 5202 \tab 1349 \tab 924 \tab 61 \tab 34 \tab 6 \tab 7576 \cr
}
}

\keyword{variable}
