% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{abpoorw}
\alias{abpoorw}
\title{Wrong for woman to get abortion if low income?}
\source{
General Social Survey https://gss.norc.org
}
\description{
abpoorw
}
\details{
Question Do you personally think it is wrong or not wrong for a woman to have an abortion. . .
B. If the family has a very low income and cannot afford any more children.
}
\section{Values}{

\itemize{
\item \verb{[1]} always wrong
\item \verb{[2]} almost always wrong
\item \verb{[3]} wrong only sometimes
\item \verb{[4]} not wrong at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4636/vshow}{GSS Data Explorer page for \code{abpoorw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always wrong \tab almost always wrong \tab wrong only sometimes \tab not wrong at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 479 \tab 195 \tab 142 \tab 408 \tab 89 \tab 46 \tab 1517 \cr
   1998 \tab 1548 \tab 549 \tab 138 \tab 112 \tab 359 \tab 94 \tab 32 \tab 2832 \cr
   2008 \tab 658 \tab 652 \tab 135 \tab 144 \tab 400 \tab 25 \tab 9 \tab 2023 \cr
   2018 \tab 1173 \tab 511 \tab 93 \tab 134 \tab 372 \tab 48 \tab 17 \tab 2348 \cr
   Total \tab 3537 \tab 2191 \tab 561 \tab 532 \tab 1539 \tab 256 \tab 104 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}
}
\concept{ISSP Religion}
\keyword{variable}
