% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{abrape}
\alias{abrape}
\title{Pregnant as result of rape}
\source{
General Social Survey https://gss.norc.org
}
\description{
abrape
}
\details{
Question Please tell me whether or not you think it should be possible for a pregnant woman to obtain a legal abortion if. . .READ EACH STATEMENT, AND CIRCLE ONE CODE FOR EACH.
E. If she became pregnant as a result of rape?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/605/vshow}{GSS Data Explorer page for \code{abrape}}.

Counts by year:\tabular{llllllll}{
   year \tab yes \tab no \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1196 \tab 316 \tab 91 \tab 1 \tab 9 \tab - \tab 1613 \cr
   1973 \tab 1212 \tab 239 \tab 50 \tab - \tab 3 \tab - \tab 1504 \cr
   1974 \tab 1228 \tab 192 \tab 62 \tab - \tab 2 \tab - \tab 1484 \cr
   1975 \tab 1190 \tab 232 \tab 65 \tab - \tab 3 \tab - \tab 1490 \cr
   1976 \tab 1205 \tab 235 \tab 52 \tab - \tab 7 \tab - \tab 1499 \cr
   1977 \tab 1231 \tab 238 \tab 52 \tab - \tab 9 \tab - \tab 1530 \cr
   1978 \tab 1232 \tab 249 \tab 45 \tab - \tab 6 \tab - \tab 1532 \cr
   1980 \tab 1177 \tab 234 \tab 54 \tab - \tab 3 \tab - \tab 1468 \cr
   1982 \tab 1485 \tab 285 \tab 81 \tab - \tab 9 \tab - \tab 1860 \cr
   1983 \tab 1246 \tab 258 \tab 63 \tab - \tab 32 \tab - \tab 1599 \cr
   1984 \tab 1130 \tab 276 \tab 57 \tab - \tab 10 \tab - \tab 1473 \cr
   1985 \tab 1197 \tab 277 \tab 57 \tab - \tab 3 \tab - \tab 1534 \cr
   1987 \tab 1376 \tab 353 \tab 79 \tab - \tab 11 \tab - \tab 1819 \cr
   1988 \tab 749 \tab 175 \tab 50 \tab 504 \tab 3 \tab - \tab 1481 \cr
   1989 \tab 825 \tab 166 \tab 37 \tab 504 \tab 5 \tab - \tab 1537 \cr
   1990 \tab 740 \tab 133 \tab 39 \tab 455 \tab 5 \tab - \tab 1372 \cr
   1991 \tab 817 \tab 128 \tab 41 \tab 524 \tab 7 \tab - \tab 1517 \cr
   1993 \tab 850 \tab 175 \tab 46 \tab 531 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 1616 \tab 318 \tab 60 \tab 996 \tab 2 \tab - \tab 2992 \cr
   1996 \tab 1550 \tab 289 \tab 77 \tab 981 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 1439 \tab 357 \tab 83 \tab 950 \tab 3 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 340 \tab 106 \tab 956 \tab 6 \tab - \tab 2817 \cr
   2002 \tab 721 \tab 184 \tab 19 \tab 1841 \tab - \tab - \tab 2765 \cr
   2004 \tab 651 \tab 203 \tab 30 \tab 1914 \tab 14 \tab - \tab 2812 \cr
   2006 \tab 1483 \tab 429 \tab 76 \tab 2507 \tab 15 \tab - \tab 4510 \cr
   2008 \tab 982 \tab 298 \tab 62 \tab 671 \tab 10 \tab - \tab 2023 \cr
   2010 \tab 987 \tab 248 \tab 28 \tab 763 \tab 18 \tab - \tab 2044 \cr
   2012 \tab 941 \tab 289 \tab 59 \tab 666 \tab 19 \tab - \tab 1974 \cr
   2014 \tab 1276 \tab 353 \tab 65 \tab 824 \tab 20 \tab - \tab 2538 \cr
   2016 \tab 1411 \tab 399 \tab 48 \tab 977 \tab 32 \tab - \tab 2867 \cr
   2018 \tab 1194 \tab 308 \tab 51 \tab 774 \tab 21 \tab - \tab 2348 \cr
   2021 \tab 1211 \tab 219 \tab 11 \tab 2574 \tab - \tab 17 \tab 4032 \cr
   2022 \tab 982 \tab 157 \tab 19 \tab 2367 \tab 13 \tab 6 \tab 3544 \cr
   Total \tab 37939 \tab 8552 \tab 1815 \tab 22280 \tab 311 \tab 23 \tab 70920 \cr
}
}

\seealso{
Other Core: 
\code{\link{abdefect}},
\code{\link{abhlth}},
\code{\link{abnomore}},
\code{\link{abpoor}},
\code{\link{absingle}},
\code{\link{fegive}},
\code{\link{fejoin}},
\code{\link{fethink}},
\code{\link{fewrite}}
}
\concept{Core}
\keyword{variable}
