% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_626.R
\name{abrapeg}
\alias{abrapeg}
\title{Pregnant as result of rape (grid on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
abrapeg
}
\details{
Question (Pleae indicate whether or not you think it should be possible for a pregnant woman to obtain a legal abortion . . .) If she becomes pregnant as a result of rape?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7470/vshow}{GSS Data Explorer page for \code{abrapeg}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab don't know \tab no answer \tab Total \cr
   2021 \tab 2822 \tab 975 \tab 204 \tab 31 \tab - \tab - \tab 4032 \cr
   2022 \tab 2380 \tab 986 \tab 135 \tab 10 \tab 26 \tab 7 \tab 3544 \cr
   Total \tab 5202 \tab 1961 \tab 339 \tab 41 \tab 26 \tab 7 \tab 7576 \cr
}
}

\keyword{variable}
