% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{abroad}
\alias{abroad}
\title{How long have you lived in other countries?}
\source{
General Social Survey https://gss.norc.org
}
\description{
abroad
}
\details{
Question About how long altogether have you lived in other countries?
}
\section{Values}{

\itemize{
\item \code{1} never lived in other countries
\item \code{2} less than 1 year in all
\item \code{3} 1 to 4 years in all
\item \code{4} 5 years or longer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4867/vshow}{GSS Data Explorer page for \code{abroad}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never lived in other countries \tab less than 1 year in all \tab 1 to 4 years in all \tab 5 years or longer \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 1014 \tab 68 \tab 119 \tab 139 \tab 1 \tab 26 \tab 2904 \cr
   Total \tab 1537 \tab 1014 \tab 68 \tab 119 \tab 139 \tab 1 \tab 26 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{refugees}},
\code{\link{res161}}
}
\concept{ISSP National Identity}
\keyword{variable}
