% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{absent}
\alias{absent}
\title{How many days absent from work past 6 month}
\source{
General Social Survey https://gss.norc.org
}
\description{
absent
}
\details{
Question About how many days have you been absent from work in the last 6 months (not counting vacation)?
}
\section{Values}{

\itemize{
\item \code{1} more thn 20 days
\item \code{2} 11 to 20 days
\item \code{3} 6 to 10 days
\item \code{4} 1 to 5 days
\item \code{5} none
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4590/vshow}{GSS Data Explorer page for \code{absent}}.
}

\seealso{
Other Work Orientation: 
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}

Other Job: 
\code{\link{edcskill}},
\code{\link{jobskill}},
\code{\link{useskill}},
\code{\link{wktenure}}

Other Work: 
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Job}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
