% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{abspno}
\alias{abspno}
\title{Abortion if husband opposes}
\source{
General Social Survey https://gss.norc.org
}
\description{
abspno
}
\details{
Question If a woman has decided to have an abortion but her husband is against it, do you think she should go ahead and have  it, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/698/vshow}{GSS Data Explorer page for \code{abspno}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 395 \tab 979 \tab 148 \tab 8 \tab 1530 \cr
   Total \tab 0 \tab 395 \tab 979 \tab 148 \tab 8 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{comprend}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{form}},
\code{\link{phone}},
\code{\link{radiohrs}}
}
\concept{Core}
\keyword{variable}
