% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{abstate1}
\alias{abstate1}
\title{Difficulty of obtaining abortion in r's state}
\source{
General Social Survey https://gss.norc.org
}
\description{
abstate1
}
\details{
Question Here in (CURRENT STATE), how easy or hard do you think it is for a woman to get an abortion?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} easy
\item \code{3} neither easy nor hard
\item \code{4} hard
\item \code{5} very hard
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7175/vshow}{GSS Data Explorer page for \code{abstate1}}.
}

\keyword{variable}
