% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{accptoth}
\alias{accptoth}
\title{R accept others even when they do things wrong}
\source{
General Social Survey https://gss.norc.org
}
\description{
accptoth
}
\details{
Question The following are things that you may experience in your daily life. Please tell me how often these occur.
B. I accept others even when they do things I think are wrong.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never or almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2012 \tab A/C/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2889/vshow}{GSS Data Explorer page for \code{accptoth}}.
}

\seealso{
Other Altruism: 
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selffrst}},
\code{\link{selfless}}

Other Feelings: 
\code{\link{selfless}}
}
\concept{Altruism}
\concept{Feelings}
\keyword{variable}
