% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_402.R
\name{acqattnd}
\alias{acqattnd}
\title{Number of regular church attendee r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqattnd
}
\details{
Question Now thinking about all the people you are acquainted with:
C. How many do you think attend religious services regularly? Would you say…
}
\section{Values}{

\itemize{
\item \code{1} almost all
\item \code{2} most
\item \code{3} about half
\item \code{4} a few
\item \code{5} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3664/vshow}{GSS Data Explorer page for \code{acqattnd}}.
}

\seealso{
Other Number Known: 
\code{\link{acqdems}},
\code{\link{acqmyrac}},
\code{\link{acqreps}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvalib}},
\code{\link{trtkaren}},
\code{\link{trtkevin}},
\code{\link{trtshawn}}
}
\concept{Number Known}
\keyword{variable}
