% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_397.R
\name{acqfmgo}
\alias{acqfmgo}
\title{Number of family member going church regularly}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqfmgo
}
\details{
Question Next, we are going to ask questions about people in your family, including relatives and in-laws.
K. How many attend religious services on a regular basis?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3612/vshow}{GSS Data Explorer page for \code{acqfmgo}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3996 \tab 37 \tab 27 \tab 130 \tab 76 \tab 207 \tab 30 \tab 7 \tab 4510 \cr
   Total \tab 3996 \tab 37 \tab 27 \tab 130 \tab 76 \tab 207 \tab 30 \tab 7 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmhsp}},
\code{\link{acqfmlib}},
\code{\link{acqfmno}},
\code{\link{acqfmwht}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}}
}
\concept{Number Known}
\keyword{variable}
