% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_397.R
\name{acqfmlib}
\alias{acqfmlib}
\title{Number of liberal in family member}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqfmlib
}
\details{
Question Next, we are going to ask questions about people in your family, including relatives and in-laws.
M. How many are strongly liberal?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3614/vshow}{GSS Data Explorer page for \code{acqfmlib}}.
}

\seealso{
Other Number Known: 
\code{\link{acqfmcoh}},
\code{\link{acqfmcon}},
\code{\link{acqfmgay}},
\code{\link{acqfmgo}},
\code{\link{acqfmhsp}},
\code{\link{acqfmno}},
\code{\link{acqfmwht}},
\code{\link{acqnhlin}},
\code{\link{acqnhmrk}}
}
\concept{Number Known}
\keyword{variable}
