% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_396.R
\name{acqmils}
\alias{acqmils}
\title{Number of ppl in military service r is acquainted with}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqmils
}
\details{
Question G. How many of the people you are acquainted with are currently serving in the armed forces?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3601/vshow}{GSS Data Explorer page for \code{acqmils}}.
}

\seealso{
Other Number Known: 
\code{\link{acqchild}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmhme}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}}

Other Social Networks: 
\code{\link{acqchild}},
\code{\link{acqelecs}},
\code{\link{acqfmasn}},
\code{\link{acqfmblk}},
\code{\link{acqfmhme}},
\code{\link{acqfmlin}},
\code{\link{acqfmmrk}},
\code{\link{acqfmpri}},
\code{\link{acqfmune}}
}
\concept{Number Known}
\concept{Social Networks}
\keyword{variable}
