% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_402.R
\name{acqmyrac}
\alias{acqmyrac}
\title{Proportion of same race as r among all acquaintances}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqmyrac
}
\details{
Question Are the people that you are acquainted with…
}
\section{Values}{

\itemize{
\item \code{1} almost all the same race as you
\item \code{2} mostly the same race as you
\item \code{3} about evenly divided
\item \code{4} mostly a different race than you
\item \code{5} almost all a different race than you
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3665/vshow}{GSS Data Explorer page for \code{acqmyrac}}.
}

\seealso{
Other Number Known: 
\code{\link{acqattnd}},
\code{\link{acqdems}},
\code{\link{acqreps}},
\code{\link{acqvacoh}},
\code{\link{acqvacon}},
\code{\link{acqvalib}},
\code{\link{trtkaren}},
\code{\link{trtkevin}},
\code{\link{trtshawn}}
}
\concept{Number Known}
\keyword{variable}
