% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_402.R
\name{acqvacon}
\alias{acqvacon}
\title{Number of conservative in voluntary association}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqvacon
}
\details{
Question Next, we are going to ask about people that you are acquainted with through schools, clubs, associations, or places of worship.
N. How many are you pretty certain are strongly conservative?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3660/vshow}{GSS Data Explorer page for \code{acqvacon}}.
}

\seealso{
Other Number Known: 
\code{\link{acqattnd}},
\code{\link{acqdems}},
\code{\link{acqmyrac}},
\code{\link{acqreps}},
\code{\link{acqvacoh}},
\code{\link{acqvalib}},
\code{\link{trtkaren}},
\code{\link{trtkevin}},
\code{\link{trtshawn}}

Other Social Networks: 
\code{\link{acqattnd}},
\code{\link{acqdems}},
\code{\link{acqmyrac}},
\code{\link{acqreps}},
\code{\link{acqvacoh}},
\code{\link{acqvalib}},
\code{\link{trtkaren}},
\code{\link{trtkevin}},
\code{\link{trtshawn}}
}
\concept{Number Known}
\concept{Social Networks}
\keyword{variable}
