% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_399.R
\name{acqwklin}
\alias{acqwklin}
\title{Number of linda in current or previous work}
\source{
General Social Survey https://gss.norc.org
}
\description{
acqwklin
}
\details{
Question Next, we are going to ask about people that you are acquainted with through work.
B. How many (IF NUMKNOWN = 2 or 5 DISPLAY: of the people you are acquainted with through work) are named Linda?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1
\item \code{3} 2 - 5
\item \code{4} 6 - 10
\item \code{5} more than 10
\item \code{6} i dont have a job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3633/vshow}{GSS Data Explorer page for \code{acqwklin}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 - 5 \tab 6 - 10 \tab more than 10 \tab don't know \tab I don't have a job \tab no answer \tab Total \cr
   2006 \tab 4009 \tab 254 \tab 118 \tab 91 \tab 12 \tab 3 \tab 12 \tab 5 \tab 6 \tab 4510 \cr
   Total \tab 4009 \tab 254 \tab 118 \tab 91 \tab 12 \tab 3 \tab 12 \tab 5 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqnhcoh}},
\code{\link{acqnhcon}},
\code{\link{acqnhlib}},
\code{\link{acqwkasn}},
\code{\link{acqwkblk}},
\code{\link{acqwkhme}},
\code{\link{acqwkmrk}},
\code{\link{acqwkpri}},
\code{\link{acqwkune}}
}
\concept{Number Known}
\keyword{variable}
