% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{activnat}
\alias{activnat}
\title{Outside leisure activity in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
activnat
}
\details{
Question In the last twelve months how often, if at all, have you engaged in any leisure activities outside in nature, such as hiking, bird watching, swimming, skiing, other outdoor activities, or just relaxing?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several times a week
\item \verb{[3]} several times a month
\item \verb{[4]} several times a year
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7496/vshow}{GSS Data Explorer page for \code{activnat}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab daily \tab several times a week \tab several times a month \tab several times a year \tab never \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 328 \tab 530 \tab 475 \tab 356 \tab 85 \tab 52 \tab 191 \tab 12 \tab 4032 \cr
   Total \tab 2003 \tab 328 \tab 530 \tab 475 \tab 356 \tab 85 \tab 52 \tab 191 \tab 12 \tab 4032 \cr
}
}

\keyword{variable}
