% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{actlaw}
\alias{actlaw}
\title{How likely r to do something if unjust law being cons}
\source{
General Social Survey https://gss.norc.org
}
\description{
actlaw
}
\details{
Question If such a case arose, how likely is it that you, acting alone or together with others, would be able to try to do something about it?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} fairly likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4942/vshow}{GSS Data Explorer page for \code{actlaw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab fairly likely \tab not very likely \tab not at all likely \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 289 \tab 589 \tab 427 \tab 146 \tab 14 \tab 7 \tab 2812 \cr
   2014 \tab 1274 \tab 193 \tab 423 \tab 379 \tab 189 \tab 77 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 482 \tab 1012 \tab 806 \tab 335 \tab 91 \tab 10 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polgreed}},
\code{\link{polint1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
