% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{actssoc}
\alias{actssoc}
\title{How well r carries out usual social activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
actssoc
}
\details{
Question In general, please rate how well you carry out your usual social activities and roles. (This includes activities at home, at work and in your community, and responsibilities as a parent, child, spouse, employee, friend, etc.)
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7167/vshow}{GSS Data Explorer page for \code{actssoc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 315 \tab 1049 \tab 743 \tab 177 \tab 41 \tab 6 \tab 17 \tab - \tab 2348 \cr
   2021 \tab 383 \tab 431 \tab 1457 \tab 1290 \tab 358 \tab 88 \tab 2 \tab 1 \tab 22 \tab 4032 \cr
   Total \tab 383 \tab 746 \tab 2506 \tab 2033 \tab 535 \tab 129 \tab 8 \tab 18 \tab 22 \tab 6380 \cr
}
}

\keyword{variable}
