% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{actupset}
\alias{actupset}
\title{Ppl at work throw things when upset with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
actupset
}
\details{
Question H. People at work throw things, slam doors, or hit objects when they are upset with me.
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3005/vshow}{GSS Data Explorer page for \code{actupset}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 31 \tab 85 \tab 170 \tab 1520 \tab 19 \tab 42 \tab 2812 \cr
   2012 \tab 795 \tab 17 \tab 52 \tab 101 \tab 966 \tab 12 \tab 31 \tab 1974 \cr
   Total \tab 1740 \tab 48 \tab 137 \tab 271 \tab 2486 \tab 31 \tab 73 \tab 4786 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
