% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{actupset}
\alias{actupset}
\title{Ppl at work throw things when upset with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
actupset
}
\details{
Question H. People at work throw things, slam doors, or hit objects when they are upset with me.
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3005/vshow}{GSS Data Explorer page for \code{actupset}}.
}

\seealso{
Other Workplace Conflict: 
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
