% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_108.R
\name{actvet}
\alias{actvet}
\title{R active in veteran group}
\source{
General Social Survey https://gss.norc.org
}
\description{
actvet
}
\details{
Question Have you ever done any active work for (EACH "YES" IN MEMFRAT-MEMOTHER)?  I mean been a leader, helped organize meetings, been an officer, or given time or money?
C. Veteran''s groups
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/998/vshow}{GSS Data Explorer page for \code{actvet}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab 1693 \tab 52 \tab 46 \tab 28 \tab 1819 \cr
   Total \tab 1693 \tab 52 \tab 46 \tab 28 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actfrat}},
\code{\link{actpolit}},
\code{\link{actserv}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solprof}}
}
\concept{Socio-political participation}
\keyword{variable}
