% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_329.R
\name{adfam}
\alias{adfam}
\title{Would r get advice from family if child was r's}
\source{
General Social Survey https://gss.norc.org
}
\description{
adfam
}
\details{
Question If (John/Mary) was your child, would you...
Get advice from family and friends about (John/Mary)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2937/vshow}{GSS Data Explorer page for \code{adfam}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfammed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldfrnd}},
\code{\link{chldsch}},
\code{\link{spendeve}}
}
\concept{Mental Health II}
\keyword{variable}
