% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{adforjob}
\alias{adforjob}
\title{Advertised for a job in newspaper's}
\source{
General Social Survey https://gss.norc.org
}
\description{
adforjob
}
\details{
Question Thinking about the last 12 months, have you done any of the following in order to find a job?
D. Advertised for a job in newspapers or journals?
}
\section{Values}{

\itemize{
\item \code{1} no
\item \code{2} yes, once or twice
\item \code{3} yes, more than twice
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4602/vshow}{GSS Data Explorer page for \code{adforjob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{econsup}},
\code{\link{rimpskls}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}

Other Unemployment: 
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{econsup}},
\code{\link{wantads}}
}
\concept{Unemployment}
\concept{Work Orientation}
\keyword{variable}
