% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{adhdcon}
\alias{adhdcon}
\title{Children should be given counseling for adhd}
\source{
General Social Survey https://gss.norc.org
}
\description{
adhdcon
}
\details{
Question Should children be given counseling for ADHD?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2971/vshow}{GSS Data Explorer page for \code{adhdcon}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{mntlrel}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}

Other Children: 
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
