% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{adhdknw}
\alias{adhdknw}
\title{What does r know about adhd (verbatim)}
\source{
General Social Survey https://gss.norc.org
}
\description{
adhdknw
}
\details{
Question PROBE: Anything else?
}
\section{Values}{

\itemize{
\item \code{1} nothing don t know
\item \code{2} had herad of it unspecified
\item \code{4} medication ritalin
\item \code{5} chemical biological cause
\item \code{6} environment social cause
\item \code{7} describing symptoms
\item \code{8} other
\item \code{30} someone else i know has it
\item \code{31} my child has it
\item \code{32} i have(had) it
\item \code{81} diet based too much sugar
\item \code{82} anti-adhd
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2969/vshow}{GSS Data Explorer page for \code{adhdknw}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{mntlrel}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}

Other Children: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
