% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{adhdreal}
\alias{adhdreal}
\title{Is adhd a real disease}
\source{
General Social Survey https://gss.norc.org
}
\description{
adhdreal
}
\details{
Question Do you think ADHD is a real disease?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2970/vshow}{GSS Data Explorer page for \code{adhdreal}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1859 \tab 685 \tab 183 \tab 37 \tab 1 \tab 2765 \cr
   Total \tab 1859 \tab 685 \tab 183 \tab 37 \tab 1 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{medsavtx}},
\code{\link{mntlrel}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
