% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{adminconsent}
\alias{adminconsent}
\title{Consent to possible data linkage}
\source{
General Social Survey https://gss.norc.org
}
\description{
adminconsent
}
\details{
Question Respondent consent to possible data linkage
}
\section{Values}{

\itemize{
\item \code{1} r consents to possible data linkage
\item \code{2} r does not consent to possible data linkage
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7425/vshow}{GSS Data Explorer page for \code{adminconsent}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab r consents to possible data linkage \tab r does not consent to possible data linkage \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 1250 \tab 1089 \tab 9 \tab - \tab 2348 \cr
   2021 \tab 549 \tab 1532 \tab 1951 \tab - \tab - \tab 4032 \cr
   2022 \tab 77 \tab 1660 \tab 1762 \tab - \tab 45 \tab 3544 \cr
   Total \tab 626 \tab 4442 \tab 4802 \tab 9 \tab 45 \tab 9924 \cr
}
}

\keyword{variable}
