% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{adoption}
\alias{adoption}
\title{Best way to match birth mothers with adopters?}
\source{
General Social Survey https://gss.norc.org
}
\description{
adoption
}
\details{
Question As you know, when a parent is unable to care for a baby, the parent often puts the baby up for adoption. For certaIn kinds of babies, the number of couples who want to adopt is much greater than the number of babies available for adoption. What do you think is the best way to match birth mothers with couples who want to adopt?
}
\section{Values}{

\itemize{
\item \code{1} a social service agency
\item \code{2} go to the highest bidders
\item \code{3} by lottery
\item \code{4} birth mothers should decide
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1869/vshow}{GSS Data Explorer page for \code{adoption}}.
}

\seealso{
Other Markets: 
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
