% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{adoptus}
\alias{adoptus}
\title{Immigrants living in us should adopt us culture}
\source{
General Social Survey https://gss.norc.org
}
\description{
adoptus
}
\details{
Question Do you think immigrants living in the U.S. should adopt U.S. culture, keep their own culture, or do both?
}
\section{Values}{

\itemize{
\item \verb{[1]} adopt u.s. culture
\item \verb{[2]} keep their own culture
\item \verb{[3]} do both
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7755/vshow}{GSS Data Explorer page for \code{adoptus}}.

Counts by year:\tabular{llllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab Total \cr
   2022 \tab 1231 \tab 258 \tab 195 \tab 1860 \tab 3544 \cr
   Total \tab 1231 \tab 258 \tab 195 \tab 1860 \tab 3544 \cr
}
}

\concept{Core}
\keyword{variable}
