% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{adoptus}
\alias{adoptus}
\title{Immigrants living in us should adopt us culture}
\source{
General Social Survey https://gss.norc.org
}
\description{
adoptus
}
\details{
Question Do you think immigrants living in the U.S. should adopt U.S. culture, keep their own culture, or do both?
}
\section{Values}{

\itemize{
\item \code{1} adopt u.s. culture
\item \code{2} keep their own culture
\item \code{3} do both
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/C \tab partial \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7755/vshow}{GSS Data Explorer page for \code{adoptus}}.
}

\concept{Core}
\keyword{variable}
