% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_330.R
\name{adpsy}
\alias{adpsy}
\title{Would r go to a psychiatrist if child was r's}
\source{
General Social Survey https://gss.norc.org
}
\description{
adpsy
}
\details{
Question If (John/Mary) was your child, would you...
Go to a psychiatrist for help?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2945/vshow}{GSS Data Explorer page for \code{adpsy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 741 \tab 605 \tab 36 \tab 11 \tab 2765 \cr
   Total \tab 1372 \tab 741 \tab 605 \tab 36 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcoumed}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsymed}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{grwingup}},
\code{\link{mntlill}}
}
\concept{Mental Health II}
\keyword{variable}
