% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_330.R
\name{adpsymed}
\alias{adpsymed}
\title{Would r have child on med by psychiatrists ad}
\source{
General Social Survey https://gss.norc.org
}
\description{
adpsymed
}
\details{
Question If (John/Mary) was your child, would you...
Go to a psychiatrist for help?
A. If the psychiatrist said your child should be on medication, how likely would you be to do that?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2946/vshow}{GSS Data Explorer page for \code{adpsymed}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adcoumed}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsy}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{grwingup}},
\code{\link{mntlill}}
}
\concept{Mental Health II}
\keyword{variable}
