% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{adults_exp}
\alias{adults_exp}
\title{Adults 18+ in hh(experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
adults_exp
}
\details{
Question Including yourself, how many individuals 18 or older live in this household?
If you have roommates, please include them. Please also include those who usually live here but are away temporarily, such as those on business trips, at school, or temporarily in a hospital.
}
\section{Values}{

\itemize{
\item \code{1} 1 adult in household
\item \code{8} 8 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7983/vshow}{GSS Data Explorer page for \code{adults_exp}}.
}

\keyword{variable}
