% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_666.R
\name{adults_exp}
\alias{adults_exp}
\title{Adults 18+ in hh(experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
adults_exp
}
\details{
Question Including yourself, how many individuals 18 or older live in this household?
If you have roommates, please include them. Please also include those who usually live here but are away temporarily, such as those on business trips, at school, or temporarily in a hospital.
}
\section{Values}{

\itemize{
\item \code{1} 1 adult in household
\item \code{8} 8 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7983/vshow}{GSS Data Explorer page for \code{adults_exp}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 adult in household \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 9 \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1788 \tab 699 \tab 764 \tab 171 \tab 63 \tab 12 \tab 6 \tab 1 \tab 1 \tab 4 \tab 35 \tab 3544 \cr
   Total \tab 1788 \tab 699 \tab 764 \tab 171 \tab 63 \tab 12 \tab 6 \tab 1 \tab 1 \tab 4 \tab 35 \tab 3544 \cr
}
}

\keyword{variable}
