% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_668.R
\name{adultsinhh}
\alias{adultsinhh}
\title{Total adults in hh (household composition version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
adultsinhh
}
\details{
Question Calculated variable: Total adults in household according to roster
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 adult in household
\item \verb{[8]} 8 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7982/vshow}{GSS Data Explorer page for \code{adultsinhh}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 adult in household \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1764 \tab 669 \tab 812 \tab 170 \tab 84 \tab 16 \tab 4 \tab 21 \tab 4 \tab 3544 \cr
   Total \tab 5796 \tab 669 \tab 812 \tab 170 \tab 84 \tab 16 \tab 4 \tab 21 \tab 4 \tab 7576 \cr
}
}

\keyword{variable}
