% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{advances}
\alias{advances}
\title{Have you advanced, same level or lost}
\source{
General Social Survey https://gss.norc.org
}
\description{
advances
}
\details{
Question Since your first full-time job with this organization, have you advanced rapidly, made steady advances, stayed at about the same level or lost some ground?
}
\section{Values}{

\itemize{
\item \code{1} advanced rapidly
\item \code{2} steady advance
\item \code{3} samelevel
\item \code{4} lost some ground
\item \code{7} 1st job with org
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1306/vshow}{GSS Data Explorer page for \code{advances}}.
}

\seealso{
Other Work Organizations: 
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}
}
\concept{Work Organizations}
\keyword{variable}
