% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_275.R
\name{advsup}
\alias{advsup}
\title{R can ask work supervisor software help}
\source{
General Social Survey https://gss.norc.org
}
\description{
advsup
}
\details{
Question Do the people you can ask for advice include? (Name as many as are correct):
H. A supervisor or trainer at work?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2506/vshow}{GSS Data Explorer page for \code{advsup}}.
}

\seealso{
Other Information Society: 
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advstu}},
\code{\link{advtchr}}

Other Computers And Internet: 
\code{\link{advcfrnd}},
\code{\link{advchld}},
\code{\link{advcowrk}},
\code{\link{advfam}},
\code{\link{advpar}},
\code{\link{advsib}},
\code{\link{advsp}},
\code{\link{advstu}},
\code{\link{advtchr}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
