% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_134.R
\name{affact}
\alias{affact}
\title{R's firm has aff. action program}
\source{
General Social Survey https://gss.norc.org
}
\description{
affact
}
\details{
Question Does the place where you work have an affirmative action program or make any special effort to hire and promote minorities?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1250/vshow}{GSS Data Explorer page for \code{affact}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1990 \tab 495 \tab 387 \tab 360 \tab 108 \tab 22 \tab 1372 \cr
   Total \tab 495 \tab 387 \tab 360 \tab 108 \tab 22 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}},
\code{\link{discwhy6}}
}
\concept{Intergroup Relations}
\keyword{variable}
