% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{afraidof}
\alias{afraidof}
\title{People act as if they are afraid of r}
\source{
General Social Survey https://gss.norc.org
}
\description{
afraidof
}
\details{
Question In your day-to-day life how often have any of the following things happened to you? People act as if they are afraid of you.
}
\section{Values}{

\itemize{
\item \code{1} almost every day
\item \code{2} at least once a week
\item \code{3} a few times a month
\item \code{4} a few times a year
\item \code{5} less than once a year
\item \code{6} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/C \tab partial \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7161/vshow}{GSS Data Explorer page for \code{afraidof}}.
}

\keyword{variable}
