% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{agape2}
\alias{agape2}
\title{Not happy unless the one s happiness is before r}
\source{
General Social Survey https://gss.norc.org
}
\description{
agape2
}
\details{
Question Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or strongly disagree with the following statements?
B. I cannot be happy unless I place the one I love''s happiness before my own.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} neither agree nor disagree
\item \code{4} disagree somewhat
\item \code{5} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3054/vshow}{GSS Data Explorer page for \code{agape2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab neither agree nor disagree \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 445 \tab 497 \tab 203 \tab 133 \tab 53 \tab 4 \tab 5 \tab 2812 \cr
   Total \tab 1472 \tab 445 \tab 497 \tab 203 \tab 133 \tab 53 \tab 4 \tab 5 \tab 2812 \cr
}
}

\seealso{
Other Altruistic Love: 
\code{\link{agape3}},
\code{\link{agape4}}
}
\concept{Altruistic Love}
\keyword{variable}
