% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{agape3}
\alias{agape3}
\title{Willing to sacrifice wish to let the one achieve}
\source{
General Social Survey https://gss.norc.org
}
\description{
agape3
}
\details{
Question Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or strongly disagree with the following statements?
C. I am usually willing to sacrifice my own wishes to let the one I love achieve his/hers.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree somewhat
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3055/vshow}{GSS Data Explorer page for \code{agape3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab neither agree nor disagree \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 488 \tab 588 \tab 160 \tab 68 \tab 26 \tab 4 \tab 6 \tab 2812 \cr
   Total \tab 1472 \tab 488 \tab 588 \tab 160 \tab 68 \tab 26 \tab 4 \tab 6 \tab 2812 \cr
}
}

\seealso{
Other Altruistic Love: 
\code{\link{agape2}},
\code{\link{agape4}}
}
\concept{Altruistic Love}
\keyword{variable}
