% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{agape4}
\alias{agape4}
\title{Would endure all things for the one r love}
\source{
General Social Survey https://gss.norc.org
}
\description{
agape4
}
\details{
Question Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or strongly disagree with the following statements?
D. I would endure all things for the sake of the one I love.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} neither agree nor disagree
\item \code{4} disagree somewhat
\item \code{5} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3056/vshow}{GSS Data Explorer page for \code{agape4}}.
}

\seealso{
Other Altruistic Love: 
\code{\link{agape2}},
\code{\link{agape3}}
}
\concept{Altruistic Love}
\keyword{variable}
