% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{agecmeus}
\alias{agecmeus}
\title{If born outside usa, what age were you when you permanently moved to usa}
\source{
General Social Survey https://gss.norc.org
}
\description{
agecmeus
}
\details{
Question If you were born outside US, what age were you when you permanently moved to the US?
}
\section{Values}{

\itemize{
\item \code{96} have not permanently moved to us
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3899/vshow}{GSS Data Explorer page for \code{agecmeus}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 46 \tab 5 \tab 52 \tab 56 \tab 6 \tab 7 \tab 9 \tab have not permanently moved to us \tab no answer \tab Total \cr
   2008 \tab 1838 \tab 9 \tab 5 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 4 \tab 7 \tab 10 \tab 8 \tab 9 \tab 9 \tab 1 \tab 7 \tab 5 \tab 7 \tab 7 \tab 7 \tab 3 \tab 3 \tab 2 \tab 4 \tab 9 \tab 2 \tab 6 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 8 \tab 6 \tab 2023 \cr
   Total \tab 1838 \tab 9 \tab 5 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 4 \tab 7 \tab 10 \tab 8 \tab 9 \tab 9 \tab 1 \tab 7 \tab 5 \tab 7 \tab 7 \tab 7 \tab 3 \tab 3 \tab 2 \tab 4 \tab 9 \tab 2 \tab 6 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 8 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{natdeg}},
\code{\link{othplan}},
\code{\link{typevisa}},
\code{\link{visa}}
}
\concept{Global Economics}
\keyword{variable}
