% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{agekdbrn}
\alias{agekdbrn}
\title{R's age when 1st child born}
\source{
General Social Survey https://gss.norc.org
}
\description{
agekdbrn
}
\details{
Question How old were you when your first child was born?
As of 2024, AGEKDBRN has been recoded to only account for first child born during ages 18-45. This has been retroactively applied to all previous years of AGEKDBRN. The original AGEKDBRN containing data from 1994 to 2022 has been renamed AGEKDBRN_9422.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/54/vshow}{GSS Data Explorer page for \code{agekdbrn}}.
}

\seealso{
Other Core: 
\code{\link{age}},
\code{\link{codeg}},
\code{\link{coeduc}},
\code{\link{degree}},
\code{\link{educ}},
\code{\link{maeduc}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{speduc}}
}
\concept{Core}
\keyword{variable}
