% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{agetech}
\alias{agetech}
\title{Are older or younger people benefiting more from digital tech}
\source{
General Social Survey https://gss.norc.org
}
\description{
agetech
}
\details{
Question Do you think that in America older people or younger people are benefitting more, or both equally, or neither, from the use of the Internet and digital technology?
}
\section{Values}{

\itemize{
\item \verb{[1]} older people are benefitting more than younger people
\item \verb{[2]} older people and younger people are benefitting equally
\item \verb{[3]} younger people are benefitting more than older people
\item \verb{[4]} neither older nor younger are benefitting
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8043/vshow}{GSS Data Explorer page for \code{agetech}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab older people are benefitting more than younger people \tab older people and younger people are benefitting equally \tab younger people are benefitting more than older people \tab neither older nor younger are benefitting \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 64 \tab 263 \tab 992 \tab 150 \tab 150 \tab 1675 \tab 4 \tab 11 \tab 3309 \cr
   Total \tab 0 \tab 64 \tab 263 \tab 992 \tab 150 \tab 150 \tab 1675 \tab 4 \tab 11 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{dataprot}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{modeorg}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
