% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{agewedcr}
\alias{agewedcr}
\title{R's age at current marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
agewedcr
}
\details{
Question How old were you when your (current) marriage began?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1481/vshow}{GSS Data Explorer page for \code{agewedcr}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 47 \tab 48 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 59 \tab 60 \tab 61 \tab 64 \tab 68 \tab 69 \tab 70 \tab 74 \tab no answer \tab Total \cr
   1994 \tab 2811 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 10 \tab 11 \tab 7 \tab 8 \tab 10 \tab 5 \tab 4 \tab 9 \tab 7 \tab 5 \tab 10 \tab 8 \tab 11 \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab 3 \tab 3 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab 2992 \cr
   Total \tab 2811 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 10 \tab 11 \tab 7 \tab 8 \tab 10 \tab 5 \tab 4 \tab 9 \tab 7 \tab 5 \tab 10 \tab 8 \tab 11 \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab 3 \tab 3 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{spfam16}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}
}
\concept{Family Mobility}
\keyword{variable}
