% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{agewedcr}
\alias{agewedcr}
\title{R's age at current marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
agewedcr
}
\details{
Question How old were you when your (current) marriage began?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1481/vshow}{GSS Data Explorer page for \code{agewedcr}}.
}

\seealso{
Other Family Mobility: 
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{spfam16}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{spmaeduc}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}
}
\concept{Family Mobility}
\keyword{variable}
