% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{aidavg}
\alias{aidavg}
\title{Gov $ for stdnts with avg inc and grades}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidavg
}
\details{
Question Some people think the government should provide financial assistance to college students. Others think the government should not provide such aid. In each of the circumstances listed below should the government provide grants that would not have to be paid back, provide loans which the student would have to pay back, or should the government not provide any financial assistance?
C. For students who have average grades and middle income parents.
}
\section{Values}{

\itemize{
\item \code{1} govt give grants
\item \code{2} govt make loans
\item \code{3} no govt asst
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4027/vshow}{GSS Data Explorer page for \code{aidavg}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}

Other Education: 
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}
}
\concept{Education}
\concept{ISSP Role of Gov't}
\keyword{variable}
