% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_444.R
\name{aidcol}
\alias{aidcol}
\title{Gov. should assist low-income college student}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidcol
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to . . .
H. Give financial assistance to college students from low-income families.
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4083/vshow}{GSS Data Explorer page for \code{aidcol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1990 \tab 155 \tab 347 \tab 638 \tab 107 \tab 34 \tab 49 \tab 42 \tab 1372 \cr
   1996 \tab 1572 \tab 428 \tab 624 \tab 134 \tab 47 \tab 44 \tab 55 \tab 2904 \cr
   2006 \tab 2992 \tab 832 \tab 559 \tab 85 \tab 36 \tab 6 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 624 \tab 583 \tab 114 \tab 42 \tab 15 \tab 12 \tab 2867 \cr
   Total \tab 6196 \tab 2231 \tab 2404 \tab 440 \tab 159 \tab 114 \tab 109 \tab 11653 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{equalize}},
\code{\link{hlthcare}},
\code{\link{protstrs}},
\code{\link{racists}},
\code{\link{revoltrs}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
