% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{aidrive}
\alias{aidrive}
\title{R's comfort with driverless car}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidrive
}
\details{
Question Various human actions can be done by robots or Artificial Intelligence (AI). For each of the following, please indicate, using a scale from 0 to 10, how you would personally feel about it. On this scale, '0' means that you would feeltotally uncomfortable and '10' means that you would feeltotally comfortable with this situation... Being driven in a driverless car.
}
\section{Values}{

\itemize{
\item \code{0} totally uncomfortable with this situation
\item \code{10} totally comfortable with this situation
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8067/vshow}{GSS Data Explorer page for \code{aidrive}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{aimed}},
\code{\link{aiworry}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
