% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsage3}
\alias{aidsage3}
\title{Age of aids victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsage3
}
\details{
Question We would like to know a few other things about that person.
C. How old (is/was) that person? (Is/Was) (he/she) . . . (READ LIST)
}
\section{Values}{

\itemize{
\item \code{1} 10 years or under
\item \code{2} 11-20
\item \code{3} 21-40
\item \code{4} 41 year or older
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5022/vshow}{GSS Data Explorer page for \code{aidsage3}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}

Other Aids: 
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}}
}
\concept{Aids}
\concept{Victims & AIDS}
\keyword{variable}
