% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{aidscndm}
\alias{aidscndm}
\title{Condom can reduce aids}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidscndm
}
\details{
Question B. Can people reduce their chance of getting the AIDS virus by using a condom every time they have sex?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5078/vshow}{GSS Data Explorer page for \code{aidscndm}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{hivkiss}},
\code{\link{hivvac}},
\code{\link{sexornt}}
}
\concept{High Risk Behaviors}
\keyword{variable}
