% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsded3}
\alias{aidsded3}
\title{Has aids victim 3rd closest to r died}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsded3
}
\details{
Question We would like to know a few other things about that person.
A. Is that person currently living, or has that person died?
}
\section{Values}{

\itemize{
\item \verb{[1]} living
\item \verb{[2]} died
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5020/vshow}{GSS Data Explorer page for \code{aidsded3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab living \tab died \tab no answer \tab don't know \tab Total \cr
   1989 \tab 1517 \tab 13 \tab 4 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 1347 \tab 11 \tab 13 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 1466 \tab 17 \tab 27 \tab 5 \tab 2 \tab 1517 \cr
   1993 \tab 1490 \tab 34 \tab 75 \tab 6 \tab 1 \tab 1606 \cr
   Total \tab 5820 \tab 75 \tab 119 \tab 15 \tab 3 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
