% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_551.R
\name{aidshlth}
\alias{aidshlth}
\title{Have govt pay for all aids health care}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidshlth
}
\details{
Question Do you support or oppose the following measures to deal with AIDS?
D. Have the government pay all of the health care costs of AIDS patients
}
\section{Values}{

\itemize{
\item \code{1} support
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5044/vshow}{GSS Data Explorer page for \code{aidshlth}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsmar}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{suiage3}},
\code{\link{suirac3}},
\code{\link{suireg3}}
}
\concept{Victims & AIDS}
\keyword{variable}
