% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{aidsknow}
\alias{aidsknow}
\title{#People r knows infected with aids}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsknow
}
\details{
Question How many people have you known personally, either living or dead, who came down with the disease called AIDS?
}
\section{Values}{

\itemize{
\item \code{7} 7 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5006/vshow}{GSS Data Explorer page for \code{aidsknow}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidswho}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex3}},
\code{\link{cidewho3}}

Other Aids: 
\code{\link{aidswho}}
}
\concept{Aids}
\concept{Victims & AIDS}
\keyword{variable}
