% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{aidslook}
\alias{aidslook}
\title{A health-look person may have aids}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidslook
}
\details{
Question A. Is it possible for a healthy-looking person to have the AIDS virus?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5077/vshow}{GSS Data Explorer page for \code{aidslook}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2014 \tab 707 \tab 1660 \tab 83 \tab 58 \tab 30 \tab 2538 \cr
   Total \tab 707 \tab 1660 \tab 83 \tab 58 \tab 30 \tab 2538 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest2}},
\code{\link{idu30}},
\code{\link{relatsex}}
}
\concept{High Risk Behaviors}
\keyword{variable}
