% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsrac3}
\alias{aidsrac3}
\title{Race of aids victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsrac3
}
\details{
Question We would like to know a few other things about that person.
D. What (is/was) that person''s race? (Is/Was) it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \code{1} black
\item \code{2} white
\item \code{3} hispanic
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5023/vshow}{GSS Data Explorer page for \code{aidsrac3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab black \tab white \tab hispanic \tab no answer \tab other \tab Total \cr
   1989 \tab 1517 \tab 1 \tab 15 \tab 1 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 1347 \tab 2 \tab 18 \tab 4 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 1466 \tab 8 \tab 34 \tab 3 \tab 6 \tab - \tab 1517 \cr
   1993 \tab 1490 \tab 20 \tab 75 \tab 10 \tab 6 \tab 5 \tab 1606 \cr
   Total \tab 5820 \tab 31 \tab 142 \tab 18 \tab 16 \tab 5 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
