% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsrac3}
\alias{aidsrac3}
\title{Race of aids victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsrac3
}
\details{
Question We would like to know a few other things about that person.
D. What (is/was) that person''s race? (Is/Was) it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \verb{[1]} black
\item \verb{[2]} white
\item \verb{[3]} hispanic
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5023/vshow}{GSS Data Explorer page for \code{aidsrac3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab black \tab white \tab hispanic \tab no answer \tab other \tab Total \cr
   1989 \tab 1517 \tab 1 \tab 15 \tab 1 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 1347 \tab 2 \tab 18 \tab 4 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 1466 \tab 8 \tab 34 \tab 3 \tab 6 \tab - \tab 1517 \cr
   1993 \tab 1490 \tab 20 \tab 75 \tab 10 \tab 6 \tab 5 \tab 1606 \cr
   Total \tab 5820 \tab 31 \tab 142 \tab 18 \tab 16 \tab 5 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
