% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsrac3}
\alias{aidsrac3}
\title{Race of aids victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsrac3
}
\details{
Question We would like to know a few other things about that person.
D. What (is/was) that person''s race? (Is/Was) it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \code{1} black
\item \code{2} white
\item \code{3} hispanic
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5023/vshow}{GSS Data Explorer page for \code{aidsrac3}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
