% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_548.R
\name{aidsrace}
\alias{aidsrace}
\title{Race of aids victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsrace
}
\details{
Question We would like to know a few other things about that person.
D. What (is/was) that person''s race? (Is/Was) it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \code{1} black
\item \code{2} white
\item \code{3} hispanic
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5011/vshow}{GSS Data Explorer page for \code{aidsrace}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab black \tab white \tab hispanic \tab other \tab no answer \tab Total \cr
   1988 \tab 1342 \tab 25 \tab 98 \tab 10 \tab 4 \tab 2 \tab 1481 \cr
   1989 \tab 1392 \tab 17 \tab 110 \tab 16 \tab - \tab 2 \tab 1537 \cr
   1990 \tab 1195 \tab 26 \tab 132 \tab 15 \tab 3 \tab 1 \tab 1372 \cr
   1991 \tab 1275 \tab 47 \tab 179 \tab 8 \tab 2 \tab 6 \tab 1517 \cr
   1993 \tab 1214 \tab 68 \tab 294 \tab 19 \tab 4 \tab 7 \tab 1606 \cr
   Total \tab 6418 \tab 183 \tab 813 \tab 68 \tab 13 \tab 18 \tab 7513 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsrac2}},
\code{\link{aidsreg}},
\code{\link{aidssex}},
\code{\link{aidssex2}},
\code{\link{aidswho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
