% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidsreg2}
\alias{aidsreg2}
\title{Region of aids victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidsreg2
}
\details{
Question We would like to know a few other things about that person.
E. What state did this person live in?
}
\section{Values}{

\itemize{
\item \code{0} foreign
\item \code{1} new england
\item \code{2} middle atlantic
\item \code{3} e. nor. central
\item \code{4} w. nor. central
\item \code{5} south atlantic
\item \code{6} e. sou. central
\item \code{7} w. sou. central
\item \code{8} mountain
\item \code{9} pacific
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5018/vshow}{GSS Data Explorer page for \code{aidsreg2}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}

Other Aids: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}}
}
\concept{Aids}
\concept{Victims & AIDS}
\keyword{variable}
