% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_548.R
\name{aidswho2}
\alias{aidswho2}
\title{R's relation. to 2nd closest aids victim}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidswho2
}
\details{
Question Think about the person you have known next best, living or dead, who came down with AIDS. Please tell me the letter of the category on the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \code{1} husband or wife
\item \code{2} partner or lover
\item \code{3} son or daughter
\item \code{4} other relative
\item \code{5} friend
\item \code{6} neighbor
\item \code{7} co-worker
\item \code{8} acquaintance
\item \code{9} patient
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5013/vshow}{GSS Data Explorer page for \code{aidswho2}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsrac2}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidssex}},
\code{\link{aidssex2}}

Other Aids: 
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsrac2}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidssex}},
\code{\link{aidssex2}}
}
\concept{Aids}
\concept{Victims & AIDS}
\keyword{variable}
