% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_548.R
\name{aidswho2}
\alias{aidswho2}
\title{R's relation. to 2nd closest aids victim}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidswho2
}
\details{
Question Think about the person you have known next best, living or dead, who came down with AIDS. Please tell me the letter of the category on the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \code{1} husband or wife
\item \code{2} partner or lover
\item \code{3} son or daughter
\item \code{4} other relative
\item \code{5} friend
\item \code{6} neighbor
\item \code{7} co-worker
\item \code{8} acquaintance
\item \code{9} patient
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5013/vshow}{GSS Data Explorer page for \code{aidswho2}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab other relative \tab friend \tab neighbor \tab co-worker \tab acquaintance \tab patient \tab other \tab no answer \tab partner or lover \tab son or daughter \tab Total \cr
   1989 \tab 1486 \tab 2 \tab 17 \tab 1 \tab 3 \tab 20 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 1537 \cr
   1990 \tab 1324 \tab 2 \tab 13 \tab 1 \tab 5 \tab 19 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 1422 \tab 5 \tab 24 \tab 7 \tab 6 \tab 30 \tab 8 \tab 8 \tab 4 \tab 1 \tab 2 \tab 1517 \cr
   1993 \tab 1407 \tab 11 \tab 51 \tab 7 \tab 29 \tab 75 \tab 9 \tab 10 \tab 7 \tab - \tab - \tab 1606 \cr
   Total \tab 5639 \tab 20 \tab 105 \tab 16 \tab 43 \tab 144 \tab 22 \tab 23 \tab 17 \tab 1 \tab 2 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage}},
\code{\link{aidsage2}},
\code{\link{aidsdead}},
\code{\link{aidsded2}},
\code{\link{aidsrac2}},
\code{\link{aidsrace}},
\code{\link{aidsreg}},
\code{\link{aidssex}},
\code{\link{aidssex2}}
}
\concept{Victims & AIDS}
\keyword{variable}
