% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{aidswho3}
\alias{aidswho3}
\title{R's relat. to 3rd closest aids victim to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
aidswho3
}
\details{
Question Think about the person you have known next best, living or dead, who came down with AIDS. Please tell me the letter of the category on the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \verb{[1]} husband or wife
\item \verb{[2]} partner or lover
\item \verb{[3]} son or daughter
\item \verb{[4]} other relative
\item \verb{[5]} friend
\item \verb{[6]} neighbor
\item \verb{[7]} co-worker
\item \verb{[8]} acquaintance
\item \verb{[9]} patient
\item \verb{[10]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5019/vshow}{GSS Data Explorer page for \code{aidswho3}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab friend \tab neighbor \tab co-worker \tab acquaintance \tab no answer \tab patient \tab other \tab son or daughter \tab other relative \tab Total \cr
   1989 \tab 1517 \tab 5 \tab 1 \tab 2 \tab 8 \tab 4 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1347 \tab 6 \tab 1 \tab 4 \tab 8 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 1466 \tab 17 \tab 1 \tab 3 \tab 17 \tab 4 \tab 6 \tab 2 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 1490 \tab 30 \tab 2 \tab 10 \tab 50 \tab 6 \tab 11 \tab 4 \tab - \tab 3 \tab 1606 \cr
   Total \tab 5820 \tab 58 \tab 5 \tab 19 \tab 83 \tab 15 \tab 20 \tab 8 \tab 1 \tab 3 \tab 6032 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{suiknew}},
\code{\link{suisex}},
\code{\link{suiwho}}
}
\concept{Victims & AIDS}
\keyword{variable}
