% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{aimofart}
\alias{aimofart}
\title{Aim of art}
\source{
General Social Survey https://gss.norc.org
}
\description{
aimofart
}
\details{
Question I'm going to read two statements about what works of art should Do. Please tell me which statement is closest to your opinion.
}
\section{Values}{

\itemize{
\item \code{1} art human spirit
\item \code{2} art xprss artist
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2128/vshow}{GSS Data Explorer page for \code{aimofart}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab art human spirit \tab art xprss artist \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 538 \tab 851 \tab 23 \tab 33 \tab 2832 \cr
   Total \tab 1387 \tab 538 \tab 851 \tab 23 \tab 33 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{drama}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{stateart}},
\code{\link{volarts}}
}
\concept{Religion and Culture}
\keyword{variable}
