% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_49.R
\name{alienat6}
\alias{alienat6}
\title{People in wash d.c. are out of touch}
\source{
General Social Survey https://gss.norc.org
}
\description{
alienat6
}
\details{
Question Now I want to read you some things some people have told us they have felt from time to time. Do you tend to feel or not . . . (READ LIST)
F. The people in Washington, D.C. are out of touch with the rest of the country.
}
\section{Values}{

\itemize{
\item \code{1} feel
\item \code{2} not feel
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/433/vshow}{GSS Data Explorer page for \code{alienat6}}.
}

\seealso{
Other Core: 
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{closeblk}},
\code{\link{closewht}},
\code{\link{hapmar}},
\code{\link{happy}}

Other Alienation: 
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat4}},
\code{\link{alienat5}}

Other Split Ballots: 
\code{\link{alienat1}},
\code{\link{alienat2}},
\code{\link{alienat3}},
\code{\link{alienat4}},
\code{\link{alienat5}},
\code{\link{hapmar}},
\code{\link{happy}}
}
\concept{Alienation}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
