% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{allben}
\alias{allben}
\title{Everyone benefit from traditional family}
\source{
General Social Survey https://gss.norc.org
}
\description{
allben
}
\details{
Question In many married couples, women take the main responsibility for the care of the home and children, while men take the main responsibility for supporting the family financially. Who benefits from this: men, women, children, everyone, or no one? PROBE: Anyone else? UNTIL R SAYS NO.
D. Everyone
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1681/vshow}{GSS Data Explorer page for \code{allben}}.
}

\seealso{
Other Gender: 
\code{\link{chldben}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{parleave}},
\code{\link{womenben}}

Other Family: 
\code{\link{chldben}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{parleave}},
\code{\link{womenben}}

Other Household: 
\code{\link{chldben}},
\code{\link{menben}},
\code{\link{womenben}}
}
\concept{Family}
\concept{Gender}
\concept{Household}
\keyword{variable}
