% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{allorglevel}
\alias{allorglevel}
\title{All levels of org are involved in stress prevention}
\source{
General Social Survey https://gss.norc.org
}
\description{
allorglevel
}
\details{
Question (To what extent to do you agree or disagree with the following statements?) In my organization, the prevention of stress involves all levels of the organization.
}
\section{Values}{

\itemize{
\item \code{1} strongly disagree
\item \code{2} disagree
\item \code{3} neither agree nor disagree
\item \code{4} agree
\item \code{5} strongly agree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7858/vshow}{GSS Data Explorer page for \code{allorglevel}}.
}

\keyword{variable}
