% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{allowmar}
\alias{allowmar}
\title{Allow racial intermarriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
allowmar
}
\details{
Question F. Do you think the law should allow marriages between blacks and whites?
}
\section{Values}{

\itemize{
\item \verb{[1]} allow marriage
\item \verb{[2]} not allow marriage
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4989/vshow}{GSS Data Explorer page for \code{allowmar}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab allow marriage \tab not allow marriage \tab don't know \tab no answer \tab Total \cr
   1989 \tab 1029 \tab 311 \tab 145 \tab 22 \tab 30 \tab 1537 \cr
   Total \tab 1029 \tab 311 \tab 145 \tab 22 \tab 30 \tab 1537 \cr
}
}

\seealso{
Other Civil Liberties: 
\code{\link{allowrac}}
}
\concept{Civil Liberties}
\keyword{variable}
