% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{allowmar}
\alias{allowmar}
\title{Allow racial intermarriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
allowmar
}
\details{
Question F. Do you think the law should allow marriages between blacks and whites?
}
\section{Values}{

\itemize{
\item \code{1} allow marriage
\item \code{2} not allow marriage
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4989/vshow}{GSS Data Explorer page for \code{allowmar}}.
}

\seealso{
Other Civil Liberties: 
\code{\link{allowrac}}
}
\concept{Civil Liberties}
\keyword{variable}
