% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{allowrac}
\alias{allowrac}
\title{Allow racist to speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
allowrac
}
\details{
Question E. Or, consider a person who believes that Blacks are genetically inferior. If such a person wanted to speech in your community claiming that Blacks are inferior, should he be allowed to speak, make a or not?
}
\section{Values}{

\itemize{
\item \code{1} yes, allow
\item \code{2} no, not allow
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4988/vshow}{GSS Data Explorer page for \code{allowrac}}.
}

\seealso{
Other Civil Liberties: 
\code{\link{allowmar}}
}
\concept{Civil Liberties}
\keyword{variable}
