% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{allowrac}
\alias{allowrac}
\title{Allow racist to speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
allowrac
}
\details{
Question E. Or, consider a person who believes that Blacks are genetically inferior. If such a person wanted to speech in your community claiming that Blacks are inferior, should he be allowed to speak, make a or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, allow
\item \verb{[2]} no, not allow
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4988/vshow}{GSS Data Explorer page for \code{allowrac}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, allow \tab no, not allow \tab don't know \tab no answer \tab Total \cr
   1989 \tab 1029 \tab 272 \tab 190 \tab 14 \tab 32 \tab 1537 \cr
   Total \tab 1029 \tab 272 \tab 190 \tab 14 \tab 32 \tab 1537 \cr
}
}

\seealso{
Other Civil Liberties: 
\code{\link{allowmar}}
}
\concept{Civil Liberties}
\keyword{variable}
