% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{altsat}
\alias{altsat}
\title{How satisfied r with alt health treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
altsat
}
\details{
Question (How satisfied or dissatisfied were you with the treatment you received...) When you last visited an alternative health care practitioner?
}
\section{Values}{

\itemize{
\item \code{1} completely satisfied
\item \code{2} very satisfied
\item \code{3} fairly satisfied
\item \code{4} neither satisfied nor dissatisfied
\item \code{5} fairly dissatisfied
\item \code{6} very dissatisfied
\item \code{7} completely dissatisfied
\item \code{8} does not apply
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7833/vshow}{GSS Data Explorer page for \code{altsat}}.
}

\keyword{variable}
