% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{amancstr}
\alias{amancstr}
\title{How important to have american ancestry}
\source{
General Social Survey https://gss.norc.org
}
\description{
amancstr
}
\details{
Question Some people say that the following things are important for being truly American. Others say they are not important. How important do you think each of the following is... (Please, select one response)
To have American ancestry
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} fairly important
\item \code{3} not very important
\item \code{4} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4881/vshow}{GSS Data Explorer page for \code{amancstr}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}

Other United States: 
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
