% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{amancstr}
\alias{amancstr}
\title{How important to have american ancestry}
\source{
General Social Survey https://gss.norc.org
}
\description{
amancstr
}
\details{
Question Some people say that the following things are important for being truly American. Others say they are not important. How important do you think each of the following is... (Please, select one response)
To have American ancestry
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} fairly important
\item \verb{[3]} not very important
\item \verb{[4]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4881/vshow}{GSS Data Explorer page for \code{amancstr}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 388 \tab 267 \tab 368 \tab 164 \tab 26 \tab 3 \tab 2812 \cr
   2014 \tab 1264 \tab 304 \tab 220 \tab 408 \tab 310 \tab 28 \tab 4 \tab 2538 \cr
   Total \tab 2860 \tab 692 \tab 487 \tab 776 \tab 474 \tab 54 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\keyword{variable}
