% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amchrstn}
\alias{amchrstn}
\title{How important to be a christian}
\source{
General Social Survey https://gss.norc.org
}
\description{
amchrstn
}
\details{
Question Some people say the following things are important for being truly American. Others say they are not important. How important do you think each of the following is...
E. To be a Christian
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} fairly important
\item \code{3} not very important
\item \code{4} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4831/vshow}{GSS Data Explorer page for \code{amchrstn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 505 \tab 198 \tab 282 \tab 324 \tab 39 \tab 19 \tab 2904 \cr
   2004 \tab 1596 \tab 585 \tab 192 \tab 209 \tab 194 \tab 35 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 408 \tab 152 \tab 271 \tab 392 \tab 44 \tab 7 \tab 2538 \cr
   Total \tab 4397 \tab 1498 \tab 542 \tab 762 \tab 910 \tab 118 \tab 27 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
