% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amcitizn}
\alias{amcitizn}
\title{Agree i would rather be a citizen of america}
\source{
General Social Survey https://gss.norc.org
}
\description{
amcitizn
}
\details{
Question How much do you agree or disagree with the following statements?
A. I would rather be a citizen of America than of any other country in the world
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4834/vshow}{GSS Data Explorer page for \code{amcitizn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 949 \tab 268 \tab 89 \tab 28 \tab 7 \tab 12 \tab 14 \tab 2904 \cr
   2004 \tab 1596 \tab 896 \tab 183 \tab 85 \tab 27 \tab 8 \tab 16 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 753 \tab 324 \tab 124 \tab 41 \tab 9 \tab 19 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 2598 \tab 775 \tab 298 \tab 96 \tab 24 \tab 47 \tab 19 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
