% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{amcult}
\alias{amcult}
\title{It is impossible to become fully american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amcult
}
\details{
Question How much do you agree or disagree with the following statements?
A. It is impossible for people who do not share American customs and traditions to become fully American
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4856/vshow}{GSS Data Explorer page for \code{amcult}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 119 \tab 329 \tab 283 \tab 430 \tab 104 \tab 46 \tab 56 \tab 2904 \cr
   2004 \tab 1596 \tab 109 \tab 281 \tab 252 \tab 437 \tab 110 \tab 26 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 58 \tab 295 \tab 194 \tab 553 \tab 122 \tab 49 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 286 \tab 905 \tab 729 \tab 1420 \tab 336 \tab 121 \tab 60 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
