% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_530.R
\name{amenglsh}
\alias{amenglsh}
\title{How important to be able to speak english}
\source{
General Social Survey https://gss.norc.org
}
\description{
amenglsh
}
\details{
Question Some people say the following things are important for being truly American. Others say they are not important. How important do you think each of the following is...
D. To be able to speak English
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} fairly important
\item \code{3} not very important
\item \code{4} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4830/vshow}{GSS Data Explorer page for \code{amenglsh}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambetter}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amshamed}},
\code{\link{belikeus}}
}
\concept{ISSP National Identity}
\keyword{variable}
