% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{amimp}
\alias{amimp}
\title{How important is being an american to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
amimp
}
\details{
Question How important is being an AMERICAN to you, where 0 is not at all important and 10 is the most important thing In your life
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{10} most
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1787/vshow}{GSS Data Explorer page for \code{amimp}}.
}

\seealso{
Other Emotions: 
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
