% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{amissue1}
\alias{amissue1}
\title{Degree r invokes primary identity as american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amissue1
}
\details{
Question D. Do you think of yourself as "just an American" on all issues, most issues, some issues, or just a few issues?
}
\section{Values}{

\itemize{
\item \code{1} all issues
\item \code{2} most issues
\item \code{3} some issues
\item \code{4} just a few issues
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1595/vshow}{GSS Data Explorer page for \code{amissue1}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{meltpot}}
}
\concept{Multiculturalism}
\keyword{variable}
