% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{amprogrn}
\alias{amprogrn}
\title{American doing enough to protect environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
amprogrn
}
\details{
Question Some countries are doing more to protect the world environment than other countries are. In general do you think that America is doing:
}
\section{Values}{

\itemize{
\item \code{1} more than enough
\item \code{2} about the right amount
\item \code{3} or, too little
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4783/vshow}{GSS Data Explorer page for \code{amprogrn}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
