% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{amproud}
\alias{amproud}
\title{R pride in being american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amproud
}
\details{
Question How proud are you to be an American . . . extremely proud, very proud, somewhat proud, or not very proud?
}
\section{Values}{

\itemize{
\item \code{1} extremely proud
\item \code{2} very proud
\item \code{3} somewhat proud
\item \code{4} not very proud
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1598/vshow}{GSS Data Explorer page for \code{amproud}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab extremely proud \tab very proud \tab somewhat proud \tab not very proud \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 683 \tab 564 \tab 185 \tab 20 \tab 7 \tab 15 \tab 2992 \cr
   Total \tab 1518 \tab 683 \tab 564 \tab 185 \tab 20 \tab 7 \tab 15 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amrank}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{meltpot}}
}
\concept{Multiculturalism}
\keyword{variable}
