% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{amproud1}
\alias{amproud1}
\title{How proud being american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amproud1
}
\details{
Question How proud are you of being American? (Please, select one response)
}
\section{Values}{

\itemize{
\item \code{1} very proud
\item \code{2} somewhat proud
\item \code{3} not very proud
\item \code{4} not proud at all
\item \code{5} i am not american
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4895/vshow}{GSS Data Explorer page for \code{amproud1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab i am not american \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 937 \tab 210 \tab 23 \tab 4 \tab 25 \tab 15 \tab 2 \tab 2812 \cr
   2014 \tab 1264 \tab 853 \tab 333 \tab 23 \tab 7 \tab 29 \tab 26 \tab 3 \tab 2538 \cr
   Total \tab 2860 \tab 1790 \tab 543 \tab 46 \tab 11 \tab 54 \tab 41 \tab 5 \tab 5350 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta2a}},
\code{\link{nafta3}}
}
\concept{ISSP National Identity}
\keyword{variable}
