% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{amsports}
\alias{amsports}
\title{Agree sports makes me proud to be an american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amsports
}
\details{
Question How much do you agree or disagree with the following statements?
F. When my country does well in international sports, it makes me proud to be an American
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4839/vshow}{GSS Data Explorer page for \code{amsports}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 485 \tab 501 \tab 257 \tab 70 \tab 18 \tab 21 \tab 15 \tab 2904 \cr
   2004 \tab 1596 \tab 415 \tab 494 \tab 246 \tab 28 \tab 14 \tab 18 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 364 \tab 583 \tab 240 \tab 49 \tab 16 \tab 19 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 1264 \tab 1578 \tab 743 \tab 147 \tab 48 \tab 58 \tab 19 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\keyword{variable}
