% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{amsports}
\alias{amsports}
\title{Agree sports makes me proud to be an american}
\source{
General Social Survey https://gss.norc.org
}
\description{
amsports
}
\details{
Question How much do you agree or disagree with the following statements?
F. When my country does well in international sports, it makes me proud to be an American
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4839/vshow}{GSS Data Explorer page for \code{amsports}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\keyword{variable}
