% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{ancestrs}
\alias{ancestrs}
\title{Believe in supernatural power of deceased ancestors}
\source{
General Social Survey https://gss.norc.org
}
\description{
ancestrs
}
\details{
Question Do you believe in…
C. ...in the supernatural powers of deceased ancestors?
}
\section{Values}{

\itemize{
\item \code{1} yes, definitely
\item \code{2} yes, probably
\item \code{3} no, probably not
\item \code{4} no, definitely not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4707/vshow}{GSS Data Explorer page for \code{ancestrs}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably not \tab no, definitely not \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 115 \tab 206 \tab 403 \tab 567 \tab 63 \tab 11 \tab 2023 \cr
   2018 \tab 1173 \tab 164 \tab 237 \tab 347 \tab 368 \tab 46 \tab 13 \tab 2348 \cr
   Total \tab 1831 \tab 279 \tab 443 \tab 750 \tab 935 \tab 109 \tab 24 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{mywaygod}},
\code{\link{nirvana}},
\code{\link{reincar}},
\code{\link{relcand}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\keyword{variable}
