% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{angrlast}
\alias{angrlast}
\title{How long did your anger last?}
\source{
General Social Survey https://gss.norc.org
}
\description{
angrlast
}
\details{
Question Emotions eventually fade away. How long did your anger or irritation last? Did it last:
}
\section{Values}{

\itemize{
\item \code{1} seconds
\item \code{2} minutes
\item \code{3} hours
\item \code{4} all that day
\item \code{5} several days, or
\item \code{6} felt continuously
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1765/vshow}{GSS Data Explorer page for \code{angrlast}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab seconds \tab minutes \tab hours \tab all that day \tab several days, or \tab felt continuously \tab no answer \tab Total \cr
   1996 \tab 1779 \tab 42 \tab 300 \tab 297 \tab 135 \tab 186 \tab 157 \tab 8 \tab 2904 \cr
   Total \tab 1779 \tab 42 \tab 300 \tab 297 \tab 135 \tab 186 \tab 157 \tab 8 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{howangry}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{thnkangr}}
}
\concept{Emotions}
\keyword{variable}
