% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_185.R
\name{angrywhy}
\alias{angrywhy}
\title{Last time r felt really angry}
\source{
General Social Survey https://gss.norc.org
}
\description{
angrywhy
}
\details{
Question Within the past month, think about the last time you felt really angry, irritated or annoyed. Could you describe in a couple of sentences what made you feel that way -- what the situation was? (INTERVIEWER: IF RESPONDENT CANNOT RECOVER AN EVENT, PROBE WITH "You haven't felt irritated or annoyed in the past month?")
D. Reasons why people get angry:
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1738/vshow}{GSS Data Explorer page for \code{angrywhy}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 23 \tab 29 \tab 3 \tab 32 \tab 4 \tab 43 \tab 44 \tab 45 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 7 \tab 8 \tab 9 \tab 90 \tab Total \cr
   1996 \tab 1779 \tab 29 \tab 88 \tab 2 \tab 4 \tab 12 \tab 11 \tab 2 \tab 6 \tab 10 \tab 11 \tab 5 \tab 45 \tab 68 \tab 36 \tab 27 \tab 78 \tab 9 \tab 3 \tab 1 \tab 9 \tab 28 \tab 15 \tab 88 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 45 \tab 60 \tab 196 \tab 173 \tab 53 \tab 2904 \cr
   Total \tab 1779 \tab 29 \tab 88 \tab 2 \tab 4 \tab 12 \tab 11 \tab 2 \tab 6 \tab 10 \tab 11 \tab 5 \tab 45 \tab 68 \tab 36 \tab 27 \tab 78 \tab 9 \tab 3 \tab 1 \tab 9 \tab 28 \tab 15 \tab 88 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 45 \tab 60 \tab 196 \tab 173 \tab 53 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywrk}},
\code{\link{beplesnt}},
\code{\link{madat1}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{whnangry}}
}
\concept{Emotions}
\keyword{variable}
