% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{anheat}
\alias{anheat}
\title{Which bodily features be best suited to a small animal living in a cold climate}
\source{
General Social Survey https://gss.norc.org
}
\description{
anheat
}
\details{
Question Please look at Card I-1. The two objects shown there have the same mass, but object B loses heat more quickly
}
\section{Values}{

\itemize{
\item \code{1} long ears and a long body
\item \code{2} small ears and a short tail
\item \code{3} a long nose and a long tail
\item \code{4} a short nose and large ears
\item \code{5} a long tail and a short nose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3472/vshow}{GSS Data Explorer page for \code{anheat}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab long ears and a long body \tab small ears and a short tail \tab a long nose and a long tail \tab a short nose and large ears \tab a long tail and a short nose \tab don't know \tab no answer \tab Total \cr
   2008 \tab 709 \tab 184 \tab 679 \tab 60 \tab 82 \tab 77 \tab 214 \tab 18 \tab 2023 \cr
   Total \tab 709 \tab 184 \tab 679 \tab 60 \tab 82 \tab 77 \tab 214 \tab 18 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
