% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{anomia7}
\alias{anomia7}
\title{Officials not interested in average man}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia7
}
\details{
Question Now I''m going to read you several more statements. (See REMARKS.)
A. Most public officials (people in public office) are not really interested in the problems of the average man.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   2006 \tab D/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/543/vshow}{GSS Data Explorer page for \code{anomia7}}.
}

\seealso{
Other Core: 
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{jobfind}},
\code{\link{jobinc}},
\code{\link{joblose}},
\code{\link{jobsec}},
\code{\link{richwork}},
\code{\link{satjob}}

Other Anomia: 
\code{\link{anomia8}},
\code{\link{anomia9}}

Other Split Ballots: 
\code{\link{anomia8}},
\code{\link{jobinc}},
\code{\link{jobsec}}
}
\concept{Anomia}
\concept{Core}
\concept{Political}
\concept{Split Ballots}
\keyword{variable}
