% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{anomia8}
\alias{anomia8}
\title{Don't know whom to trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia8
}
\details{
Question Now I''m going to read you several more statements. (See REMARKS.)
B. These days a person doesn''t really know whom he can count on.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab - \tab full \cr
   1974 \tab - \tab full \cr
   1976 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/544/vshow}{GSS Data Explorer page for \code{anomia8}}.
}

\seealso{
Other Core: 
\code{\link{anomia7}},
\code{\link{anomia9}},
\code{\link{jobfind}},
\code{\link{jobinc}},
\code{\link{joblose}},
\code{\link{jobsec}},
\code{\link{richwork}},
\code{\link{satjob}}

Other Anomia: 
\code{\link{anomia7}},
\code{\link{anomia9}}

Other Split Ballots: 
\code{\link{anomia7}},
\code{\link{jobinc}},
\code{\link{jobsec}}
}
\concept{Anomia}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
