% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{anscitst}
\alias{anscitst}
\title{Scientists shld be allowed to do research animals' pain & injury}
\source{
General Social Survey https://gss.norc.org
}
\description{
anscitst
}
\details{
Question Scientists should be allowed to do research that causes pain and injury to animals like dogs and chimpanzees if it produces new information about human health problems. (Do you strongly agree, agree, disagree, or strongly disagree?)
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3469/vshow}{GSS Data Explorer page for \code{anscitst}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}

Other Science: 
\code{\link{anheat}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Environment}
\concept{Health}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
