% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{antirel}
\alias{antirel}
\title{Anti-religious material should be banned}
\source{
General Social Survey https://gss.norc.org
}
\description{
antirel
}
\details{
Question Some books or films offend people who have strong religious beliefs. Should books and films that attack religions be prohibited by law or should they be allowed?
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4684/vshow}{GSS Data Explorer page for \code{antirel}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{socright}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
