% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{antiwar}
\alias{antiwar}
\title{Anti-war demonstration}
\source{
General Social Survey https://gss.norc.org
}
\description{
antiwar
}
\details{
Question Have you ever taken part in:
C. An anti-war demonstration?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/655/vshow}{GSS Data Explorer page for \code{antiwar}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab Total \cr
   1973 \tab - \tab 73 \tab 1430 \tab 1 \tab 1504 \cr
   Total \tab 0 \tab 73 \tab 1430 \tab 1 \tab 1504 \cr
}
}

\seealso{
Other Core: 
\code{\link{civright}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitnum}},
\code{\link{prowar}},
\code{\link{school}},
\code{\link{strike}}
}
\concept{Core}
\keyword{variable}
