% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{arrest}
\alias{arrest}
\title{Ever picked up or charged by police}
\source{
General Social Survey https://gss.norc.org
}
\description{
arrest
}
\details{
Question A. Were you ever picked up, or charged, by the police, for any (other) reason whether or not you were guilty?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
   Full \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/685/vshow}{GSS Data Explorer page for \code{arrest}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab refused \tab no answer \tab don't know \tab Total \cr
   1973 \tab - \tab 156 \tab 1288 \tab 1 \tab 59 \tab - \tab 1504 \cr
   1974 \tab - \tab 148 \tab 1328 \tab - \tab 8 \tab - \tab 1484 \cr
   1976 \tab - \tab 128 \tab 1337 \tab - \tab 33 \tab 1 \tab 1499 \cr
   1977 \tab - \tab 156 \tab 1341 \tab - \tab 33 \tab - \tab 1530 \cr
   1980 \tab - \tab 178 \tab 1244 \tab 1 \tab 42 \tab 3 \tab 1468 \cr
   1982 \tab - \tab 233 \tab 1603 \tab 1 \tab 22 \tab 1 \tab 1860 \cr
   1984 \tab - \tab 186 \tab 1265 \tab - \tab 21 \tab 1 \tab 1473 \cr
   2012 \tab 220 \tab 373 \tab 1375 \tab - \tab 5 \tab 1 \tab 1974 \cr
   Total \tab 220 \tab 1558 \tab 10781 \tab 3 \tab 223 \tab 7 \tab 12792 \cr
}
}

\seealso{
Other Core: 
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}},
\code{\link{tvhours}}
}
\concept{Core}
\keyword{variable}
