% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_235.R
\name{artatt}
\alias{artatt}
\title{Attended exhibit alone or with others}
\source{
General Social Survey https://gss.norc.org
}
\description{
artatt
}
\details{
Question You said that in the last 12 months you attended an exhibit. I'd like to ask you a few questions about the last time you did this type of activity.
Did you attend this exhibit alone, with one other person, or with two or more other people?
}
\section{Values}{

\itemize{
\item \code{1} alone
\item \code{2} with one other person
\item \code{3} with two or more other people
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6231/vshow}{GSS Data Explorer page for \code{artatt}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt4}},
\code{\link{artatt5}},
\code{\link{artfree}},
\code{\link{artwhy1}},
\code{\link{artwhy2}},
\code{\link{prfmwhy0}}
}
\concept{Arts and Culture}
\keyword{variable}
