% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_233.R
\name{artexbt}
\alias{artexbt}
\title{Did r go to an art exhibit in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
artexbt
}
\details{
Question During the last 12 months, did you go to an art exhibit, such as paintings, sculpture, textiles, graphic design, or photography?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6210/vshow}{GSS Data Explorer page for \code{artexbt}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 481 \tab 988 \tab 8 \tab 2867 \cr
   Total \tab 1390 \tab 481 \tab 988 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{prfmatt}},
\code{\link{prfmatt1}},
\code{\link{prfmatt2}},
\code{\link{prfmatt3}},
\code{\link{prfmatt4}},
\code{\link{prfmatt5}},
\code{\link{prfmdan}},
\code{\link{prfmmus}},
\code{\link{prfmthe}}
}
\concept{Arts and Culture}
\keyword{variable}
