% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{artgod}
\alias{artgod}
\title{Thru art and music we come closer to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
artgod
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example . . .
A. Through such things as art and music, we come closer to God.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2124/vshow}{GSS Data Explorer page for \code{artgod}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{artists}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{natarts}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}

Other Religion: 
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion}
\concept{Religion and Culture}
\keyword{variable}
