% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{arthrtis}
\alias{arthrtis}
\title{Told have arthritis or rheumatism}
\source{
General Social Survey https://gss.norc.org
}
\description{
arthrtis
}
\details{
Question Now I would like to ask you some questions about general health conditions. Has a doctor, nurse, or other health professional EVER told you that you had:
B. Arthritis or rheumatism?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2843/vshow}{GSS Data Explorer page for \code{arthrtis}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{knowschd}},
\code{\link{satjob1}},
\code{\link{slpprblm}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}

Other Health: 
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{slpprblm}},
\code{\link{weight}}
}
\concept{Health}
\concept{Quality of Working Life}
\keyword{variable}
