% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{artinfo}
\alias{artinfo}
\title{Use web to find info about artists}
\source{
General Social Survey https://gss.norc.org
}
\description{
artinfo
}
\details{
Question E. Visit the Web to get information about an artist or sculptor or a school of artists?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2477/vshow}{GSS Data Explorer page for \code{artinfo}}.
}

\seealso{
Other Information Society: 
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{musicnew}},
\code{\link{wwwart}},
\code{\link{wwwworld}}
}
\concept{Information Society}
\keyword{variable}
