% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{artists}
\alias{artists}
\title{Excellent artists able support themselves}
\source{
General Social Survey https://gss.norc.org
}
\description{
artists
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example . . .
B. Artists and performers who are truly excellent will be able to support themselves by selling artworks or tickets to performances.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2125/vshow}{GSS Data Explorer page for \code{artists}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 108 \tab 697 \tab 385 \tab 68 \tab 174 \tab 13 \tab 2832 \cr
   Total \tab 1387 \tab 108 \tab 697 \tab 385 \tab 68 \tab 174 \tab 13 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{artgod}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{natarts}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion and Culture}
\keyword{variable}
