% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{artloc}
\alias{artloc}
\title{Did not want to go to that location}
\source{
General Social Survey https://gss.norc.org
}
\description{
artloc
}
\details{
Question Did not want to go to that location.
}
\section{Values}{

\itemize{
\item \code{1} a reason
\item \code{2} not a reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6261/vshow}{GSS Data Explorer page for \code{artloc}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{artgst}},
\code{\link{artmostoth}},
\code{\link{artothr}},
\code{\link{arttime}},
\code{\link{prfmmostoth}}
}
\concept{Arts and Culture}
\keyword{variable}
