% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{artmostoth}
\alias{artmostoth}
\title{Most imp factor in not attending art exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
artmostoth
}
\details{
Question Of these factors, which one was the most important in your decision not to attend?
}
\section{Values}{

\itemize{
\item \code{1} cost too much
\item \code{2} programs or events were not of interest
\item \code{3} too difficult to get there
\item \code{4} could not find anyone to go with
\item \code{5} could not find the time
\item \code{6} did not want to go to that location
\item \code{7} work conflict
\item \code{8} ill, too tired
\item \code{10} no childcare
\item \code{11} sold out
\item \code{12} forgot, didn't act in time
\item \code{13} away, other conflict
\item \code{14} taking care of family member
\item \code{20} some other reason not shown
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6265/vshow}{GSS Data Explorer page for \code{artmostoth}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{artgst}},
\code{\link{artloc}},
\code{\link{artothr}},
\code{\link{arttime}},
\code{\link{prfmmostoth}}
}
\concept{Arts and Culture}
\keyword{variable}
