% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{artmostoth}
\alias{artmostoth}
\title{Most imp factor in not attending art exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
artmostoth
}
\details{
Question Of these factors, which one was the most important in your decision not to attend?
}
\section{Values}{

\itemize{
\item \verb{[1]} cost too much
\item \verb{[2]} programs or events were not of interest
\item \verb{[3]} too difficult to get there
\item \verb{[4]} could not find anyone to go with
\item \verb{[5]} could not find the time
\item \verb{[6]} did not want to go to that location
\item \verb{[7]} work conflict
\item \verb{[8]} ill, too tired
\item \verb{[10]} no childcare
\item \verb{[11]} sold out
\item \verb{[12]} forgot, didn't act in time
\item \verb{[13]} away, other conflict
\item \verb{[14]} taking care of family member
\item \verb{[20]} some other reason not shown
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6265/vshow}{GSS Data Explorer page for \code{artmostoth}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab cost too much \tab programs or events were not of interest \tab too difficult to get there \tab could not find anyone to go with \tab could not find the time \tab did not want to go to that location \tab work conflict \tab ill, too tired \tab no childcare \tab away, other conflict \tab taking care of family member \tab some other reason not shown \tab no answer \tab Total \cr
   2016 \tab 2780 \tab 19 \tab 2 \tab 10 \tab 5 \tab 26 \tab 1 \tab 1 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2867 \cr
   Total \tab 2780 \tab 19 \tab 2 \tab 10 \tab 5 \tab 26 \tab 1 \tab 1 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artgst}},
\code{\link{artloc}},
\code{\link{artothr}},
\code{\link{arttime}},
\code{\link{prfmmostoth}}
}
\concept{Arts and Culture}
\keyword{variable}
