% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{artshr}
\alias{artshr}
\title{How many hours at arts websites}
\source{
General Social Survey https://gss.norc.org
}
\description{
artshr
}
\details{
Question In a typical month, about how many minutes or hours do you use the Web to explore sites that have to do with music, film, the visual arts, dance, theater, museums, literature or related topics?
B. Hours per month?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2452/vshow}{GSS Data Explorer page for \code{artshr}}.
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
