% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{artsnew}
\alias{artsnew}
\title{R learn something new on web about arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
artsnew
}
\details{
Question Now I would like you to think about the last time you used the Web to get information or exchange views about music, film, the visual arts, dance, theater, museums, literature, or related topics.
A. On that occasion, did you come across or learn about anything new with regard to any of these topics?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[8]} cant recall
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2454/vshow}{GSS Data Explorer page for \code{artsnew}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 2325 \tab 213 \tab 131 \tab 30 \tab 118 \tab 2817 \cr
   2002 \tab 2284 \tab 239 \tab 196 \tab - \tab 46 \tab 2765 \cr
   Total \tab 4609 \tab 452 \tab 327 \tab 30 \tab 164 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
