% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{artview}
\alias{artview}
\title{Use web to look at art works}
\source{
General Social Survey https://gss.norc.org
}
\description{
artview
}
\details{
Question D. Visit a Web site that is not associated with an art museum or art gallery in order to look at art works?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2476/vshow}{GSS Data Explorer page for \code{artview}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2604 \tab 68 \tab 30 \tab 115 \tab 2817 \cr
   2002 \tab 2628 \tab 56 \tab 34 \tab 47 \tab 2765 \cr
   Total \tab 5232 \tab 124 \tab 64 \tab 162 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{musicnew}},
\code{\link{wwwart}},
\code{\link{wwwworld}}
}
\concept{Information Society}
\keyword{variable}
