% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_235.R
\name{artwhy1}
\alias{artwhy1}
\title{Imp of low cost in decision to attend exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
artwhy1
}
\details{
Question People attend exhibits for different reasons. How big a reason was low cost in your decision to attend this exhibit? Was it a major reason, a minor reason, or not a reason at all?     
SAY, "THIS CARD SHOWS SOME REASONS THAT PEOPLE ATTEND EXHIBITS."
}
\section{Values}{

\itemize{
\item \code{1} major reason
\item \code{2} minor reason
\item \code{3} not a reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6238/vshow}{GSS Data Explorer page for \code{artwhy1}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{artatt}},
\code{\link{artatt1}},
\code{\link{artatt2}},
\code{\link{artatt3}},
\code{\link{artatt4}},
\code{\link{artatt5}},
\code{\link{artfree}},
\code{\link{artwhy2}},
\code{\link{prfmwhy0}}
}
\concept{Arts and Culture}
\keyword{variable}
