% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_236.R
\name{artwhy4}
\alias{artwhy4}
\title{Imp of wanting to celebrate cultural heritage in decision to attend exhibit}
\source{
General Social Survey https://gss.norc.org
}
\description{
artwhy4
}
\details{
Question How big a reason was wanting to learn about or celebrate your or your family's cultural heritage in your decision to attend this exhibit? Was it a major reason, a minor reason, or not a reason at all?
}
\section{Values}{

\itemize{
\item \code{1} major reason
\item \code{2} minor reason
\item \code{3} not a reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6241/vshow}{GSS Data Explorer page for \code{artwhy4}}.
}

\seealso{
Other Arts and Culture: 
\code{\link{artwhy3}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo}},
\code{\link{nogo1}},
\code{\link{prfmcost}}
}
\concept{Arts and Culture}
\keyword{variable}
