% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{askcomp}
\alias{askcomp}
\title{Govt investigate use of computers}
\source{
General Social Survey https://gss.norc.org
}
\description{
askcomp
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
J. Illegal or unauthorized use of computers
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5217/vshow}{GSS Data Explorer page for \code{askcomp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 1053 \tab 247 \tab 32 \tab 32 \tab 45 \tab 10 \tab 2817 \cr
   Total \tab 1398 \tab 1053 \tab 247 \tab 32 \tab 32 \tab 45 \tab 10 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{secterr}}
}
\concept{National Security}
\keyword{variable}
