% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{askdrugs}
\alias{askdrugs}
\title{Govt investigate workers illegal drug use}
\source{
General Social Survey https://gss.norc.org
}
\description{
askdrugs
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
C. Illegal drug use
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should
\item \verb{[2]} probably should
\item \verb{[3]} probably should not
\item \verb{[4]} definitely should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5210/vshow}{GSS Data Explorer page for \code{askdrugs}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 1234 \tab 169 \tab 28 \tab 16 \tab 15 \tab 12 \tab 2992 \cr
   1996 \tab 1444 \tab 1246 \tab 148 \tab 31 \tab 12 \tab 19 \tab 4 \tab 2904 \cr
   1998 \tab 1445 \tab 1101 \tab 212 \tab 28 \tab 14 \tab 23 \tab 9 \tab 2832 \cr
   2000 \tab 1398 \tab 1143 \tab 189 \tab 28 \tab 19 \tab 32 \tab 8 \tab 2817 \cr
   Total \tab 5805 \tab 4724 \tab 718 \tab 115 \tab 61 \tab 89 \tab 33 \tab 11545 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{secdocs}},
\code{\link{sectech}}
}
\concept{National Security}
\keyword{variable}
