% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_290.R
\name{askelse}
\alias{askelse}
\title{R asked someone else for help}
\source{
General Social Survey https://gss.norc.org
}
\description{
askelse
}
\details{
Question When you want advice or assistance with surfing on the Web, finding a Web site, downloading software or some other online service, or getting your Internet connection to work, which of the following do you do? (Name as many as are correct.)
E. Ask someone else you know for help (like a friend or family member)
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2627/vshow}{GSS Data Explorer page for \code{askelse}}.
}

\seealso{
Other Information Society: 
\code{\link{askwksch}},
\code{\link{browser1}},
\code{\link{browser2}},
\code{\link{cussup}},
\code{\link{doself}},
\code{\link{manbook}},
\code{\link{payelse}},
\code{\link{processr}},
\code{\link{tecsup}}
}
\concept{Information Society}
\keyword{variable}
