% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{askfinan}
\alias{askfinan}
\title{Govt investigate workers finances for security}
\source{
General Social Survey https://gss.norc.org
}
\description{
askfinan
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
A. Financial and credit history
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5208/vshow}{GSS Data Explorer page for \code{askfinan}}.
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{secdocs}},
\code{\link{sectech}}

Other United States: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{secdocs}},
\code{\link{sectech}}
}
\concept{National Security}
\concept{United States}
\keyword{variable}
