% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{askfrbiz}
\alias{askfrbiz}
\title{Govt investigate foreign business contacts}
\source{
General Social Survey https://gss.norc.org
}
\description{
askfrbiz
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to ask him or her detailed, personal questions in the following areas:
H. Foreign business contacts
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should
\item \verb{[2]} probably should
\item \verb{[3]} probably should not
\item \verb{[4]} definitely should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5215/vshow}{GSS Data Explorer page for \code{askfrbiz}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 846 \tab 371 \tab 90 \tab 49 \tab 53 \tab 10 \tab 2817 \cr
   Total \tab 1398 \tab 846 \tab 371 \tab 90 \tab 49 \tab 53 \tab 10 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{secterr}}
}
\concept{National Security}
\keyword{variable}
