% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{askhelp}
\alias{askhelp}
\title{Asked relatives friends colleagues to help}
\source{
General Social Survey https://gss.norc.org
}
\description{
askhelp
}
\details{
Question Thinking about the last 12 months, have you done any of the following in order to find a job?
F. Asked relatives, friends, or colleagues to help you find a job?
}
\section{Values}{

\itemize{
\item \verb{[1]} no
\item \verb{[2]} yes, once or twice
\item \verb{[3]} yes, more than twice
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4604/vshow}{GSS Data Explorer page for \code{askhelp}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes, once or twice \tab yes, more than twice \tab no answer \tab don't know \tab Total \cr
   1998 \tab 2460 \tab 251 \tab 41 \tab 19 \tab 61 \tab - \tab 2832 \cr
   2006 \tab 4008 \tab 405 \tab 48 \tab 44 \tab 4 \tab 1 \tab 4510 \cr
   2016 \tab 2533 \tab 227 \tab 45 \tab 39 \tab 23 \tab - \tab 2867 \cr
   Total \tab 9001 \tab 883 \tab 134 \tab 102 \tab 88 \tab 1 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{econsup}},
\code{\link{rimpskls}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}
}
\concept{Work Orientation}
\keyword{variable}
