% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{astrolgy}
\alias{astrolgy}
\title{Ever read a horoscope or personal astrology report}
\source{
General Social Survey https://gss.norc.org
}
\description{
astrolgy
}
\details{
Question Now, for a new subject. Do you ever read a horoscope or your personal astrology report?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3323/vshow}{GSS Data Explorer page for \code{astrolgy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1073 \tab 789 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 518 \tab 811 \tab 691 \tab 1 \tab 2 \tab 2023 \cr
   2010 \tab 1353 \tab 367 \tab 322 \tab - \tab 2 \tab 2044 \cr
   2012 \tab 972 \tab 492 \tab 504 \tab 2 \tab 4 \tab 1974 \cr
   2014 \tab 1299 \tab 551 \tab 687 \tab - \tab 1 \tab 2538 \cr
   2016 \tab 1477 \tab 616 \tab 770 \tab 1 \tab 3 \tab 2867 \cr
   2018 \tab 1173 \tab 506 \tab 666 \tab 2 \tab 1 \tab 2348 \cr
   Total \tab 9438 \tab 4416 \tab 4429 \tab 8 \tab 13 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{odds1}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
