% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_580.R
\name{attreg}
\alias{attreg}
\title{Last 7 days attend a worship service}
\source{
General Social Survey https://gss.norc.org
}
\description{
attreg
}
\details{
Question During the last seven days did you do the following:
A. Attend a regular, weekly worship service at a church/synagogue (e.g. mass or Sunday morning services). Don't include watching a service on TV or listening to one on the radio.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5318/vshow}{GSS Data Explorer page for \code{attreg}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2608 \tab 219 \tab 33 \tab 44 \tab 2904 \cr
   1998 \tab 2328 \tab 452 \tab 45 \tab 7 \tab 2832 \cr
   Total \tab 4936 \tab 671 \tab 78 \tab 51 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{friday}},
\code{\link{mediarel}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
