% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{attweek}
\alias{attweek}
\title{Attend religious events during the last week}
\source{
General Social Survey https://gss.norc.org
}
\description{
attweek
}
\details{
Question Did you, yourself, happen to attend church or synagogue in the last seven days?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5324/vshow}{GSS Data Explorer page for \code{attweek}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2395 \tab 193 \tab 313 \tab 3 \tab 2904 \cr
   Total \tab 2395 \tab 193 \tab 313 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
