% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{autonojb}
\alias{autonojb}
\title{R heard of persons in r's firm having jobs replaced by computers}
\source{
General Social Survey https://gss.norc.org
}
\description{
autonojb
}
\details{
Question Have you heard of persons in your firm having their jobs replaced by computers, computerized equipment or other Forms of automation in the past three years?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3889/vshow}{GSS Data Explorer page for \code{autonojb}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 1381 \tab 101 \tab 541 \tab 2023 \cr
   Total \tab 1381 \tab 101 \tab 541 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
