% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{autonojb}
\alias{autonojb}
\title{R heard of persons in r's firm having jobs replaced by computers}
\source{
General Social Survey https://gss.norc.org
}
\description{
autonojb
}
\details{
Question Have you heard of persons in your firm having their jobs replaced by computers, computerized equipment or other Forms of automation in the past three years?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3889/vshow}{GSS Data Explorer page for \code{autonojb}}.
}

\seealso{
Other Global Economics: 
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}

Other Computers And Internet: 
\code{\link{wocomptr}}

Other Employment: 
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{whynojb1}},
\code{\link{whynojb2}}

Other Job: 
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}

Other Technology: 
\code{\link{meetf2f1}},
\code{\link{meetf2f2}}

Other Unemployment: 
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{whynojb1}},
\code{\link{whynojb2}}

Other Work: 
\code{\link{meetf2f1}},
\code{\link{meetf2f2}}
}
\concept{Business}
\concept{Computers And Internet}
\concept{Corporations}
\concept{Employment}
\concept{Global Economics}
\concept{Job}
\concept{Occupation}
\concept{Technology}
\concept{Unemployment}
\concept{Work}
\keyword{variable}
