% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_539.R
\name{avoidbuy}
\alias{avoidbuy}
\title{Boycotted products for pol reasons}
\source{
General Social Survey https://gss.norc.org
}
\description{
avoidbuy
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the past year, whether you have done it in the more distant past, whether you have not done it but might do it, or have not done it and would never, under any circumstances, do it.
B. Boycotted, or deliberately bought, certain products for political, ethical or environmental reasons
}
\section{Values}{

\itemize{
\item \code{1} have done it in the past yr
\item \code{2} have done it in the more distant past
\item \code{3} have not done it but might do it
\item \code{4} have not done it and would never do it
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4921/vshow}{GSS Data Explorer page for \code{avoidbuy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab have done it in the past yr \tab have done it in the more distant past \tab have not done it but might do it \tab have not done it and would never do it \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 351 \tab 215 \tab 488 \tab 403 \tab 11 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 294 \tab 164 \tab 358 \tab 393 \tab 52 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 645 \tab 379 \tab 846 \tab 796 \tab 63 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{attrally}},
\code{\link{cntctgov}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{polfunds}},
\code{\link{racmeet}},
\code{\link{revmeet}},
\code{\link{signdpet}},
\code{\link{usemedia}}
}
\concept{ISSP Citizenship}
\keyword{variable}
