% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_393.R
\name{avoidmh}
\alias{avoidmh}
\title{R changed shopping place due to presence of mentally ill ppl}
\source{
General Social Survey https://gss.norc.org
}
\description{
avoidmh
}
\details{
Question Have you changed where you go for shopping or entertainment because of the presence of people who seem to be mentally ill…
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} almost never
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3570/vshow}{GSS Data Explorer page for \code{avoidmh}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab almost never \tab never \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3073 \tab 18 \tab 70 \tab 175 \tab 1156 \tab 6 \tab 12 \tab 4510 \cr
   Total \tab 3073 \tab 18 \tab 70 \tab 175 \tab 1156 \tab 6 \tab 12 \tab 4510 \cr
}
}

\seealso{
Other Mental Health III: 
\code{\link{gesttalk}}
}
\concept{Mental Health III}
\keyword{variable}
