% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_393.R
\name{avoidmh}
\alias{avoidmh}
\title{R changed shopping place due to presence of mentally ill ppl}
\source{
General Social Survey https://gss.norc.org
}
\description{
avoidmh
}
\details{
Question Have you changed where you go for shopping or entertainment because of the presence of people who seem to be mentally ill…
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} almost never
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3570/vshow}{GSS Data Explorer page for \code{avoidmh}}.
}

\seealso{
Other Mental Health III: 
\code{\link{gesttalk}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
