% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{babies_exp}
\alias{babies_exp}
\title{Kids 0-6 in hh(experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
babies_exp
}
\details{
Question Household Size and Composition (see Appendix D:  Recodes, for information about these variables).  A. NUMBER OF MEMBERS UNDER 6 YEARS OLD:
As of 2024, BABIES_EXP has been recoded to only account for 0 to 2+ children younger than 6 in the household. This has been retroactively applied to all previous years of BABIES_EXP. The original BABIES_EXP containing data in 2022 has been renamed BABIES_EXP_22.
}
\section{Values}{

\itemize{
\item \verb{[0]} 0 household members under 6
\item \verb{[8]} 8 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7998/vshow}{GSS Data Explorer page for \code{babies_exp}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 0 household members under 6 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1788 \tab 1457 \tab 178 \tab 73 \tab 22 \tab 8 \tab 1 \tab 9 \tab 8 \tab 3544 \cr
   Total \tab 1788 \tab 1457 \tab 178 \tab 73 \tab 22 \tab 8 \tab 1 \tab 9 \tab 8 \tab 3544 \cr
}
}

\seealso{
Other Core: 
\code{\link{hompop_exp}}
}
\concept{Core}
\keyword{variable}
