% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{backpain}
\alias{backpain}
\title{R had back pain in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
backpain
}
\details{
Question A. In the past 12 months, have you had back pain every day for a week or more?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2831/vshow}{GSS Data Explorer page for \code{backpain}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 506 \tab 1270 \tab 5 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 475 \tab 1248 \tab 1 \tab 10 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 297 \tab 863 \tab 1 \tab 26 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 285 \tab 960 \tab - \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 378 \tab 1036 \tab - \tab 34 \tab - \tab 2348 \cr
   2022 \tab 1558 \tab 542 \tab 1401 \tab 3 \tab 17 \tab 23 \tab 3544 \cr
   Total \tab 8348 \tab 2483 \tab 6778 \tab 10 \tab 107 \tab 23 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{strredpg}},
\code{\link{usedup}}
}
\concept{Quality of Working Life}
\keyword{variable}
