% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{backpain}
\alias{backpain}
\title{R had back pain in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
backpain
}
\details{
Question A. In the past 12 months, have you had back pain every day for a week or more?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2831/vshow}{GSS Data Explorer page for \code{backpain}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{strredpg}},
\code{\link{usedup}}

Other Health: 
\code{\link{hlthdays}},
\code{\link{painarms}},
\code{\link{physhlth}},
\code{\link{usedup}}
}
\concept{Health}
\concept{Quality of Working Life}
\keyword{variable}
