% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{ballot}
\alias{ballot}
\title{Ballot used for interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
ballot
}
\details{
Question Ballot Administered.
}
\section{Values}{

\itemize{
\item \verb{[1]} ballot a
\item \verb{[2]} ballot b
\item \verb{[3]} ballot c
\item \verb{[4]} ballot d
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{version}}
}
\concept{Sampling and Weights}
\keyword{variable}
