% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{ballot}
\alias{ballot}
\title{Ballot used for interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
ballot
}
\details{
Question Ballot Administered.
}
\section{Values}{

\itemize{
\item \code{1} ballot a
\item \code{2} ballot b
\item \code{3} ballot c
\item \code{4} ballot d
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5574/vshow}{GSS Data Explorer page for \code{ballot}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{version}}

Other Sample: 
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{version}}
}
\concept{Sample}
\concept{Sampling and Weights}
\keyword{variable}
