% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{barate}
\alias{barate}
\title{Rate of ba degree earning in college respondent attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
barate
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
C. Graduation Rate
}
\section{Values}{

\itemize{
\item \verb{[1]} 0-39.9pct
\item \verb{[2]} 40-57.9pct
\item \verb{[3]} 58pct or over
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/71/vshow}{GSS Data Explorer page for \code{barate}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 0-39.9pct \tab 40-57.9pct \tab 58pct or over \tab no answer \tab Total \cr
   2012 \tab 1483 \tab 59 \tab 183 \tab 246 \tab 3 \tab 1974 \cr
   2014 \tab 1877 \tab 89 \tab 212 \tab 356 \tab 4 \tab 2538 \cr
   2016 \tab 2075 \tab 93 \tab 229 \tab 463 \tab 7 \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 7783 \tab 241 \tab 624 \tab 1065 \tab 14 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
