% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{batch}
\alias{batch}
\title{Sample batch}
\source{
General Social Survey https://gss.norc.org
}
\description{
batch
}
\details{
Question Sample batch
}
\section{Values}{

\itemize{
\item \code{1} batch 1
\item \code{2} batch 2
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8038/vshow}{GSS Data Explorer page for \code{batch}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{subsamprate}}
}
\concept{Sampling and Weights}
\keyword{variable}
