% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{beausprt}
\alias{beausprt}
\title{Spiritually touched by beauty creation}
\source{
General Social Survey https://gss.norc.org
}
\description{
beausprt
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
F. I am spiritually touched by the beauty of creation.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never/almst nevr
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2113/vshow}{GSS Data Explorer page for \code{beausprt}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never/almst nevr \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 357 \tab 401 \tab 245 \tab 176 \tab 135 \tab 86 \tab 13 \tab 32 \tab 2832 \cr
   2004 \tab 1472 \tab 287 \tab 441 \tab 244 \tab 166 \tab 97 \tab 82 \tab 7 \tab 16 \tab 2812 \cr
   Total \tab 2859 \tab 644 \tab 842 \tab 489 \tab 342 \tab 232 \tab 168 \tab 20 \tab 48 \tab 5644 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{effort}},
\code{\link{fidgety}},
\code{\link{godlove}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{uniongod}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
