% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{befair}
\alias{befair}
\title{How often do you think people take advantage}
\source{
General Social Survey https://gss.norc.org
}
\description{
befair
}
\details{
Question How often do you think that people would try to take advantage of you if they got the chance and how often would they try to be fair?
}
\section{Values}{

\itemize{
\item \code{1} try to take advantage almost all of the time
\item \code{2} try to take advantage most of the time
\item \code{3} try to be fair most of the time
\item \code{4} try to be fair almost all of the time
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4685/vshow}{GSS Data Explorer page for \code{befair}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{socright}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
