% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{believe}
\alias{believe}
\title{Importance of believing in god w o doubt}
\source{
General Social Survey https://gss.norc.org
}
\description{
believe
}
\details{
Question People have many differing views about what makes a person a good Christian or Jew. Please tell me how important is each of the following to you. Once again, please indicate where you would place your feelings on a scale from 1 to 5. How important is it …
B. To believe in God without question or doubt.
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{5} not very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1090/vshow}{GSS Data Explorer page for \code{believe}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab 2 \tab 3 \tab 4 \tab not very important \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 782 \tab 271 \tab 179 \tab 94 \tab 132 \tab 11 \tab 12 \tab 1481 \cr
   1998 \tab 1387 \tab 841 \tab 212 \tab 172 \tab 67 \tab 120 \tab 21 \tab 12 \tab 2832 \cr
   Total \tab 1387 \tab 1623 \tab 483 \tab 351 \tab 161 \tab 252 \tab 32 \tab 24 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
