% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{belvedoc}
\alias{belvedoc}
\title{R believe what doctor said}
\source{
General Social Survey https://gss.norc.org
}
\description{
belvedoc
}
\details{
Question Do you believe what the doctor just told you?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} maybe
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2752/vshow}{GSS Data Explorer page for \code{belvedoc}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab maybe \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 1552 \tab 797 \tab 390 \tab 21 \tab 5 \tab 2765 \cr
   Total \tab 0 \tab 1552 \tab 797 \tab 390 \tab 21 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
