% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_185.R
\name{beplesnt}
\alias{beplesnt}
\title{I try to be pleasant so not to upset oth.}
\source{
General Social Survey https://gss.norc.org
}
\description{
beplesnt
}
\details{
Question The following are some statements that people have made. For each one, please tell me if you strongly agree, agree, disagree, or strongly disagree with the statement. (NOTE: INTERVIEWERS SHOULD CODE "neither agree or disagree/it depends" ONLY WHEN VOLUNTEERED.)
I. I try to be pleasant so that others won't get upset.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree or disagree-it depends
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1733/vshow}{GSS Data Explorer page for \code{beplesnt}}.
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{angrlast}},
\code{\link{angry}},
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{anxious}},
\code{\link{ashamed}},
\code{\link{atease}},
\code{\link{badbrks}},
\code{\link{calm}},
\code{\link{chnang1}},
\code{\link{chnang10}},
\code{\link{chnang11}},
\code{\link{chnang12}},
\code{\link{chnang13}},
\code{\link{chnang14}},
\code{\link{chnang15}},
\code{\link{chnang16}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{chnang6}},
\code{\link{chnang7}},
\code{\link{chnang8}},
\code{\link{chnang9}},
\code{\link{contentd}},
\code{\link{embarrss}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{hapfeel}},
\code{\link{howangry}},
\code{\link{likeoth}},
\code{\link{litcntrl}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{madat1}},
\code{\link{madat10}},
\code{\link{madat11}},
\code{\link{madat12}},
\code{\link{madat13}},
\code{\link{madat14}},
\code{\link{madat15}},
\code{\link{madat16}},
\code{\link{madat17}},
\code{\link{madat18}},
\code{\link{madat19}},
\code{\link{madat2}},
\code{\link{madat20}},
\code{\link{madat21}},
\code{\link{madat22}},
\code{\link{madat23}},
\code{\link{madat3}},
\code{\link{madat4}},
\code{\link{madat5}},
\code{\link{madat6}},
\code{\link{madat7}},
\code{\link{madat8}},
\code{\link{madat9}},
\code{\link{mostluck}},
\code{\link{noemote}},
\code{\link{noplan}},
\code{\link{notam}},
\code{\link{notupset}},
\code{\link{notworry}},
\code{\link{othresp}},
\code{\link{outraged}},
\code{\link{ovrjoyed}},
\code{\link{proud}},
\code{\link{reactok}},
\code{\link{restless}},
\code{\link{sad}},
\code{\link{selfresp}},
\code{\link{shakeblu}},
\code{\link{showangr}},
\code{\link{showfeel}},
\code{\link{thnkangr}},
\code{\link{usangry}},
\code{\link{usenthus}},
\code{\link{usfrustr}},
\code{\link{ushopefl}},
\code{\link{ussat}},
\code{\link{usupset}},
\code{\link{usworry}},
\code{\link{whnangry}},
\code{\link{worried}}
}
\concept{Emotions}
\keyword{variable}
