% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{bestfrd}
\alias{bestfrd}
\title{Number of other close friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
bestfrd
}
\details{
Question Now think about your best friend, the friend you feel closest to (but not your partner). Is this best friend ...
}
\section{Values}{

\itemize{
\item \verb{[1]} a male relative
\item \verb{[2]} a female relative
\item \verb{[3]} a man who is not a relative
\item \verb{[4]} a woman who is not a relative
\item \verb{[5]} i dont have a close friend
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4204/vshow}{GSS Data Explorer page for \code{bestfrd}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab a male relative \tab a female relative \tab a man who is not a relative \tab a woman who is not a relative \tab i dont have a close friend \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 151 \tab 186 \tab 328 \tab 419 \tab 56 \tab 2 \tab 7 \tab 2765 \cr
   Total \tab 1616 \tab 151 \tab 186 \tab 328 \tab 419 \tab 56 \tab 2 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
