% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{bestfrd}
\alias{bestfrd}
\title{Number of other close friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
bestfrd
}
\details{
Question Now think about your best friend, the friend you feel closest to (but not your partner). Is this best friend ...
}
\section{Values}{

\itemize{
\item \code{1} a male relative
\item \code{2} a female relative
\item \code{3} a man who is not a relative
\item \code{4} a woman who is not a relative
\item \code{5} i dont have a close friend
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4204/vshow}{GSS Data Explorer page for \code{bestfrd}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
