% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{betrlang}
\alias{betrlang}
\title{Which language r speaks more fluent}
\source{
General Social Survey https://gss.norc.org
}
\description{
betrlang
}
\details{
Question B. Which language do you speak more fluent?
}
\section{Values}{

\itemize{
\item \code{1} language 1
\item \code{2} language 2
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2302/vshow}{GSS Data Explorer page for \code{betrlang}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{othhome}},
\code{\link{spklangw}},
\code{\link{uswht}}

Other Language: 
\code{\link{othhome}},
\code{\link{spklangw}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
