% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{biasnews}
\alias{biasnews}
\title{News coverage about politics in america is biased}
\source{
General Social Survey https://gss.norc.org
}
\description{
biasnews
}
\details{
Question To what extent do you agree or disagree with the following statement?… Media coverage about politics in America is biased.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8045/vshow}{GSS Data Explorer page for \code{biasnews}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 584 \tab 610 \tab 243 \tab 62 \tab 13 \tab 106 \tab 1667 \tab 6 \tab 18 \tab 3309 \cr
   Total \tab 0 \tab 584 \tab 610 \tab 243 \tab 62 \tab 13 \tab 106 \tab 1667 \tab 6 \tab 18 \tab 3309 \cr
}
}

\concept{ISSP National Identity}
\keyword{variable}
