% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{biasnews}
\alias{biasnews}
\title{News coverage about politics in america is biased}
\source{
General Social Survey https://gss.norc.org
}
\description{
biasnews
}
\details{
Question To what extent do you agree or disagree with the following statement?… Media coverage about politics in America is biased.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8045/vshow}{GSS Data Explorer page for \code{biasnews}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{bizmoney}},
\code{\link{conpow}},
\code{\link{experts}},
\code{\link{fulldem}},
\code{\link{incgap1}}
}
\concept{ISSP National Identity}
\keyword{variable}
