% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{bible}
\alias{bible}
\title{Feelings about the bible}
\source{
General Social Survey https://gss.norc.org
}
\description{
bible
}
\details{
Question A. Which of these statements comes closest to describing your feelings about the Bible?  READ FIRST THREE STATEMENTS ((a)-(c)) ONLY. CIRCLE ONLY ONE CODE.
a.         The Bible is the actual word of God and is to be taken literally, word for word .....1nb. The Bible is the inspired word of God but not everything in it should be taken literally, word for word ..... .2nc.The Bible is an ancient book of fables, legends, history, and moral precepts recorded by men ......3
}
\section{Values}{

\itemize{
\item \code{1} word of god
\item \code{2} inspired word
\item \code{3} ancient book
\item \code{4} other (vol)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/364/vshow}{GSS Data Explorer page for \code{bible}}.
}

\seealso{
Other Core: 
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}

Other Religion: 
\code{\link{bibley}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}

Other Split Ballots: 
\code{\link{bibley}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}}
}
\concept{Core}
\concept{Religion}
\concept{Split Ballots}
\keyword{variable}
