% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{bible1}
\alias{bible1}
\title{Describe your feelings about the bible}
\source{
General Social Survey https://gss.norc.org
}
\description{
bible1
}
\details{
Question Which of these statements comes closest to describing your feelings about the Bible?
}
\section{Values}{

\itemize{
\item \code{1} word of god, take literally
\item \code{2} god-inspired, shouldn't take literally
\item \code{3} book of fables, by man
\item \code{4} this does not apply to me
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4656/vshow}{GSS Data Explorer page for \code{bible1}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
