% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{biblev}
\alias{biblev}
\title{Feelings about the bible (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
biblev
}
\details{
Question Which of these statements comes closest to describing your feelings about the Bible?
A.) The Bible is the actual word of God and is to be taken literally, word for word.
B.) The Bible is the inspired word of God but not everything in it should be taken
literally, word for word.
C.) The Bible is an ancient book of fables, legends, history, and moral precepts recorded by man.
}
\section{Values}{

\itemize{
\item \code{1} word of god
\item \code{2} inspired word
\item \code{3} ancient book
\item \code{4} other (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7597/vshow}{GSS Data Explorer page for \code{biblev}}.
}

\keyword{variable}
