% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{biblev}
\alias{biblev}
\title{Feelings about the bible (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
biblev
}
\details{
Question Which of these statements comes closest to describing your feelings about the Bible?
A.) The Bible is the actual word of God and is to be taken literally, word for word.
B.) The Bible is the inspired word of God but not everything in it should be taken
literally, word for word.
C.) The Bible is an ancient book of fables, legends, history, and moral precepts recorded by man.
}
\section{Values}{

\itemize{
\item \verb{[1]} word of god
\item \verb{[2]} inspired word
\item \verb{[3]} ancient book
\item \verb{[4]} other (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7597/vshow}{GSS Data Explorer page for \code{biblev}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab word of god \tab inspired word \tab ancient book \tab other (phone mode only: volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2007 \tab 424 \tab 816 \tab 574 \tab 186 \tab 3 \tab 3 \tab 19 \tab 4032 \cr
   2022 \tab 2698 \tab 163 \tab 349 \tab 223 \tab 105 \tab - \tab - \tab 6 \tab 3544 \cr
   Total \tab 4705 \tab 587 \tab 1165 \tab 797 \tab 291 \tab 3 \tab 3 \tab 25 \tab 7576 \cr
}
}

\keyword{variable}
