% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_545.R
\name{big5d2}
\alias{big5d2}
\title{R see oneself as someone who gets nervous easily}
\source{
General Social Survey https://gss.norc.org
}
\description{
big5d2
}
\details{
Question To what extent do you agree or disagree with the following statements? I see myself as someone who…
D. …gets nervous easily
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4982/vshow}{GSS Data Explorer page for \code{big5d2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 73 \tab 325 \tab 192 \tab 694 \tab 232 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 73 \tab 325 \tab 192 \tab 694 \tab 232 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Big 5: 
\code{\link{big5a2}},
\code{\link{big5b2}},
\code{\link{big5c2}},
\code{\link{big5e1}},
\code{\link{big5e2}}
}
\concept{Big 5}
\keyword{variable}
