% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{bigbang1}
\alias{bigbang1}
\title{Universe began with a big explosion: true or false}
\source{
General Social Survey https://gss.norc.org
}
\description{
bigbang1
}
\details{
Question According to astronomers, the universe began with a huge explosion. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/- \tab 2016, 2018 \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3548/vshow}{GSS Data Explorer page for \code{bigbang1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 287 \tab 85 \tab 109 \tab 4 \tab 1974 \cr
   2016 \tab 2388 \tab 320 \tab 80 \tab 75 \tab 4 \tab 2867 \cr
   2018 \tab 1958 \tab 255 \tab 83 \tab 52 \tab - \tab 2348 \cr
   Total \tab 5835 \tab 862 \tab 248 \tab 236 \tab 8 \tab 7189 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bigbang}},
\code{\link{bigbang2}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{hotcore}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{radioact}},
\code{\link{viruses}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
