% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{bilinged}
\alias{bilinged}
\title{Support or oppose bilingual education}
\source{
General Social Survey https://gss.norc.org
}
\description{
bilinged
}
\details{
Question How do you feel about bilingual education? Are you strongly in favor of it, somewhat in favor of it, somewhat opposed to it, or strongly opposed to it?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} somewhat favor
\item \code{3} somewhat opposed
\item \code{4} strongly opposed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1607/vshow}{GSS Data Explorer page for \code{bilinged}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}

Other Education: 
\code{\link{engteach}}

Other Language: 
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}}
}
\concept{Education}
\concept{Language}
\concept{Multiculturalism}
\keyword{variable}
