% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{biosci}
\alias{biosci}
\title{How scientific: biology}
\source{
General Social Survey https://gss.norc.org
}
\description{
biosci
}
\details{
Question Please look at Card B20. How scientific are each of the following fields? If you have not heard of a particular field, just say you haven’t heard of it
E. Biology. Is biology very scientific, pretty scientific, not too scientific, or not scientific at all?
}
\section{Values}{

\itemize{
\item \code{1} very scientific
\item \code{2} pretty scientific
\item \code{3} not too scientific
\item \code{4} not scientific at all
\item \code{5} haven't heard of it (volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3410/vshow}{GSS Data Explorer page for \code{biosci}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{medsci}}

Other Science: 
\code{\link{accntsci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{medsci}}
}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
