% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{bizbstgw}
\alias{bizbstgw}
\title{Best policy vs narrow interest: business leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizbstgw
}
\details{
Question Please look at Card B5. When making policy recommendations about global warming, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole versus what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
C. Business Leaders
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3364/vshow}{GSS Data Explorer page for \code{bizbstgw}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizinfgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}

Other Environment: 
\code{\link{bizinfgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}}

Other Science: 
\code{\link{bizinfgw}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Environment}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
