% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_374.R
\name{bizbsttx}
\alias{bizbsttx}
\title{Best policy vs narrow interest on fed inc tax: business leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizbsttx
}
\details{
Question Please look at Card B14. When making policy recommendations about federal income taxes, on a scale of 1 to 5, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests?
INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
B. Business leaders
}
\section{Values}{

\itemize{
\item \code{1} what is best for the country
\item \code{5} own narrow interests
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3385/vshow}{GSS Data Explorer page for \code{bizbsttx}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizinftx}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{polbsttx}},
\code{\link{polinftx}},
\code{\link{txpol}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
