% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{bizgross}
\alias{bizgross}
\title{Business revenue last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizgross
}
\details{
Question In which of these groups did the revenues from this business fall for last year
}
\section{Values}{

\itemize{
\item \code{1} ( $15,000
\item \code{2} $15,001-$25,000
\item \code{3} $25,001-$30,000
\item \code{4} $30,001-$35,000
\item \code{5} $35,001-$50,000
\item \code{6} $50,001-$60,000
\item \code{7} $60,001-$75,000
\item \code{8} $75,001-$100,000
\item \code{9} ) $100,000
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3819/vshow}{GSS Data Explorer page for \code{bizgross}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab ( $15,000 \tab $15,001-$25,000 \tab $25,001-$30,000 \tab $30,001-$35,000 \tab $35,001-$50,000 \tab $50,001-$60,000 \tab $60,001-$75,000 \tab $75,001-$100,000 \tab ) $100,000 \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1773 \tab 104 \tab 16 \tab 11 \tab 7 \tab 17 \tab 9 \tab 7 \tab 7 \tab 48 \tab 12 \tab 12 \tab 2023 \cr
   Total \tab 1773 \tab 104 \tab 16 \tab 11 \tab 7 \tab 17 \tab 9 \tab 7 \tab 7 \tab 48 \tab 12 \tab 12 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}}
}
\concept{Self Employment}
\keyword{variable}
