% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{bizgross}
\alias{bizgross}
\title{Business revenue last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizgross
}
\details{
Question In which of these groups did the revenues from this business fall for last year
}
\section{Values}{

\itemize{
\item \verb{[1]} ( $15,000
\item \verb{[2]} $15,001-$25,000
\item \verb{[3]} $25,001-$30,000
\item \verb{[4]} $30,001-$35,000
\item \verb{[5]} $35,001-$50,000
\item \verb{[6]} $50,001-$60,000
\item \verb{[7]} $60,001-$75,000
\item \verb{[8]} $75,001-$100,000
\item \verb{[9]} ) $100,000
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3819/vshow}{GSS Data Explorer page for \code{bizgross}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab ( $15,000 \tab $15,001-$25,000 \tab $25,001-$30,000 \tab $30,001-$35,000 \tab $35,001-$50,000 \tab $50,001-$60,000 \tab $60,001-$75,000 \tab $75,001-$100,000 \tab ) $100,000 \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1773 \tab 104 \tab 16 \tab 11 \tab 7 \tab 17 \tab 9 \tab 7 \tab 7 \tab 48 \tab 12 \tab 12 \tab 2023 \cr
   Total \tab 1773 \tab 104 \tab 16 \tab 11 \tab 7 \tab 17 \tab 9 \tab 7 \tab 7 \tab 48 \tab 12 \tab 12 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}}
}
\concept{Self Employment}
\keyword{variable}
