% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{bizgross}
\alias{bizgross}
\title{Business revenue last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizgross
}
\details{
Question In which of these groups did the revenues from this business fall for last year
}
\section{Values}{

\itemize{
\item \code{1} ( $15,000
\item \code{2} $15,001-$25,000
\item \code{3} $25,001-$30,000
\item \code{4} $30,001-$35,000
\item \code{5} $35,001-$50,000
\item \code{6} $50,001-$60,000
\item \code{7} $60,001-$75,000
\item \code{8} $75,001-$100,000
\item \code{9} ) $100,000
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3819/vshow}{GSS Data Explorer page for \code{bizgross}}.
}

\seealso{
Other Self Employment: 
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{numown}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{startbiz}},
\code{\link{whybiz}},
\code{\link{yearbiz}}
}
\concept{Self Employment}
\keyword{variable}
