% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_372.R
\name{bizinfgw}
\alias{bizinfgw}
\title{Influence on global warming policy: business leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizinfgw
}
\details{
Question Please look at Card B4. How much influence should each of the following groups have in deciding what  to do about global warming?
C. Business Leaders. Would you say a great deal of influence, a fair amount, a little influence, or none at all?
}
\section{Values}{

\itemize{
\item \code{1} a great deal of influence
\item \code{2} a fair amount
\item \code{3} a little influence
\item \code{4} none at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3361/vshow}{GSS Data Explorer page for \code{bizinfgw}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstgw}},
\code{\link{gasregs}},
\code{\link{polbstgw}},
\code{\link{polinfgw}},
\code{\link{scibstgw}},
\code{\link{sciinfgw}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
