% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{bizinfnk}
\alias{bizinfnk}
\title{Business leaders how much influence}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizinfnk
}
\details{
Question Some say that the government should expand the use of nuclear power to generate electricity in the U.S. Others oppose any expansion in the use of nuclear power in this country. Please look at Card E15. How much influence should each of the following groups have in deciding whether to expand the use of nuclear power?
Business leaders. (Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} a great deal of influence
\item \code{2} a fair amount
\item \code{3} a little influence
\item \code{4} none at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3494/vshow}{GSS Data Explorer page for \code{bizinfnk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal of influence \tab a fair amount \tab a little influence \tab none at all \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1578 \tab 22 \tab 111 \tab 200 \tab 106 \tab 20 \tab 7 \tab 2044 \cr
   Total \tab 1578 \tab 22 \tab 111 \tab 200 \tab 106 \tab 20 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstnk}},
\code{\link{biznuke}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{nukeelec}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
