% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{biznuke}
\alias{biznuke}
\title{How well business leader know risk by nuclear power}
\source{
General Social Survey https://gss.norc.org
}
\description{
biznuke
}
\details{
Question Here's another public issue: nuclear power. Nuclear power uses the energy of the atom to generate electricity. some people say that nuclear power is a safe option to meet the energy needs of the United States without burning fossil fuels such as coal, oil, or natural gas. Others are concerned about potential risks in the operation of nuclear plants and about the disposal of nuclear waste. Please look at Card E13. (On a scale of 1 to 5, where 1 means "Very Well" and 5 means  Not At All), how well do the following groups understand the likely effects of using nuclear power to generate electricity?
B. Business leaders (How well business leaders know risk by nuclear power?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very well
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3489/vshow}{GSS Data Explorer page for \code{biznuke}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very well \tab 2 \tab 3 \tab 4 \tab not at all \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1578 \tab 16 \tab 56 \tab 140 \tab 142 \tab 85 \tab 20 \tab 7 \tab 2044 \cr
   Total \tab 1578 \tab 16 \tab 56 \tab 140 \tab 142 \tab 85 \tab 20 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstnk}},
\code{\link{bizinfnk}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{nukeelec}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
