% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{bizpeop}
\alias{bizpeop}
\title{Business people do more for environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizpeop
}
\details{
Question A. On balance, which of these two do you think is making more effort to look after the environment:
}
\section{Values}{

\itemize{
\item \verb{[1]} business and industry
\item \verb{[2]} or people in general
\item \verb{[3]} both equally
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4784/vshow}{GSS Data Explorer page for \code{bizpeop}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab business and industry \tab or people in general \tab both equally \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 143 \tab 555 \tab 296 \tab 178 \tab 104 \tab 2817 \cr
   Total \tab 1541 \tab 143 \tab 555 \tab 296 \tab 178 \tab 104 \tab 2817 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
