% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{bizshare}
\alias{bizshare}
\title{Percentage ownership r holds}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizshare
}
\details{
Question What percent ownership do you currently hold in this business?
IF R HAS MORE THAN ONE BUSINESS/FARM/PARTNERSHIP, ASK ABOUT THE PRIMARY BUSINESS THEY ARE INVOLVED IN. ALLOW THE CAPTURE OF NON-NUMERICAL ANSWERS LIKE "MAJORITY", "CONTROLLING
INTEREST", "ONE OF THREE PARTNERS", ETC.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3820/vshow}{GSS Data Explorer page for \code{bizshare}}.
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{numemps5}},
\code{\link{numown}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{startbiz}},
\code{\link{whybiz}},
\code{\link{yearbiz}}
}
\concept{Self Employment}
\keyword{variable}
