% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{bizshare}
\alias{bizshare}
\title{Percentage ownership r holds}
\source{
General Social Survey https://gss.norc.org
}
\description{
bizshare
}
\details{
Question What percent ownership do you currently hold in this business?
IF R HAS MORE THAN ONE BUSINESS/FARM/PARTNERSHIP, ASK ABOUT THE PRIMARY BUSINESS THEY ARE INVOLVED IN. ALLOW THE CAPTURE OF NON-NUMERICAL ANSWERS LIKE "MAJORITY", "CONTROLLING
INTEREST", "ONE OF THREE PARTNERS", ETC.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3820/vshow}{GSS Data Explorer page for \code{bizshare}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 15 \tab 20 \tab 25 \tab 30 \tab 33 \tab 41 \tab 5 \tab 50 \tab 51 \tab 60 \tab 70 \tab 90 \tab don't know \tab Total \cr
   2008 \tab 1773 \tab 10 \tab 2 \tab 1 \tab 164 \tab 1 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab 44 \tab 4 \tab 2 \tab 1 \tab 1 \tab 6 \tab 2023 \cr
   Total \tab 1773 \tab 10 \tab 2 \tab 1 \tab 164 \tab 1 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab 44 \tab 4 \tab 2 \tab 1 \tab 1 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{numemps5}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}}
}
\concept{Self Employment}
\keyword{variable}
