% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{blkgains}
\alias{blkgains}
\title{Best way for blacks to improve position}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkgains
}
\details{
Question People have different ideas about the best way for blacks to improve their position in American life and politics. Some people think that the best way for blacks to improve their position is through civil rights groups, they would be at point 1 on this card. Other people think that the best for way blacks to improve their position is for each individual black to become better trained and more qualified. They would be located at point 7. And other people have opinions somewhere in between. Where would you place yourself on this scale?
}
\section{Values}{

\itemize{
\item \code{1} civil rights grps
\item \code{7} better trained inds
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1042/vshow}{GSS Data Explorer page for \code{blkgains}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{blkinflu}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{gavepol}}
}
\concept{Socio-political participation}
\keyword{variable}
