% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{blkgains}
\alias{blkgains}
\title{Best way for blacks to improve position}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkgains
}
\details{
Question People have different ideas about the best way for blacks to improve their position in American life and politics. Some people think that the best way for blacks to improve their position is through civil rights groups, they would be at point 1 on this card. Other people think that the best for way blacks to improve their position is for each individual black to become better trained and more qualified. They would be located at point 7. And other people have opinions somewhere in between. Where would you place yourself on this scale?
}
\section{Values}{

\itemize{
\item \verb{[1]} civil rights grps
\item \verb{[7]} better trained inds
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1042/vshow}{GSS Data Explorer page for \code{blkgains}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab civil rights grps \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab better trained inds \tab no answer \tab Total \cr
   1987 \tab 1275 \tab 37 \tab 19 \tab 17 \tab 140 \tab 72 \tab 41 \tab 209 \tab 9 \tab 1819 \cr
   Total \tab 1275 \tab 37 \tab 19 \tab 17 \tab 140 \tab 72 \tab 41 \tab 209 \tab 9 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{blkinflu}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{gavepol}}
}
\concept{Socio-political participation}
\keyword{variable}
