% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_175.R
\name{blkgovt}
\alias{blkgovt}
\title{Do blacks get proper govt attention}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkgovt
}
\details{
Question B. Do you think that blacks get more attention from government than they deserve?  Would you answer much more attention from government than they deserve, more attention than they deserve, about the right amount of attention, less attention than they deserve, or much less attention from government than they deserve?
}
\section{Values}{

\itemize{
\item \code{1} much more
\item \code{2} more
\item \code{3} about right
\item \code{4} less
\item \code{5} much less
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1633/vshow}{GSS Data Explorer page for \code{blkgovt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much more \tab more \tab about right \tab less \tab much less \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 230 \tab 404 \tab 415 \tab 298 \tab 73 \tab 45 \tab 9 \tab 2992 \cr
   Total \tab 1518 \tab 230 \tab 404 \tab 415 \tab 298 \tab 73 \tab 45 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{asngovt}},
\code{\link{econpast}},
\code{\link{hspgovt}},
\code{\link{obrespct}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Multiculturalism}
\keyword{variable}
