% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{blkinflu}
\alias{blkinflu}
\title{Opinion of political influence of blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkinflu
}
\details{
Question People have different opinions about the amount of influence that various groups have in American life and politics. Do you think that blacks have far too much influence, too much influence, about the right amount of influence, too little influence, or do they far too little influence?
}
\section{Values}{

\itemize{
\item \code{1} far too much
\item \code{2} too much
\item \code{3} right amount
\item \code{4} too little
\item \code{5} far too little
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1041/vshow}{GSS Data Explorer page for \code{blkinflu}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{blkgains}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{gavepol}}

Other Race: 
\code{\link{blkgains}}
}
\concept{Race}
\concept{Socio-political participation}
\keyword{variable}
