% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_132.R
\name{blkschs}
\alias{blkschs}
\title{Spend more on black schools}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkschs
}
\details{
Question Here are several things that the government in Washington might Do to deal with the problems of poverty and
unemployment among Black Americans.
I would like you to tell me if you favor or oppose them. (READ "A") - Would you say that you strongly favor it, favor it, neither favor it nor oppose it, oppose it, or strongly oppose it?
B. Spending more money on the schools in black neighborhoods, especially for pre-school and early education programs.
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1234/vshow}{GSS Data Explorer page for \code{blkschs}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{blkcol}},
\code{\link{blkzone}},
\code{\link{influblk}},
\code{\link{influjew}},
\code{\link{influwht}},
\code{\link{povcol}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}}
}
\concept{Intergroup Relations}
\keyword{variable}
