% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_657.R
\name{blktry}
\alias{blktry}
\title{Black people do not try hard enough}
\source{
General Social Survey https://gss.norc.org
}
\description{
blktry
}
\details{
Question (Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or disagree strongly with the following statements?) It's really a matter of some people not trying hard enough; if black people would only try harder they could be

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} neither agree nor disagree
\item \code{4} disagree somewhat
\item \code{5} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7895/vshow}{GSS Data Explorer page for \code{blktry}}.
}

\seealso{
Other GSS Next: 
\code{\link{blkdsrv}},
\code{\link{brv5}},
\code{\link{brv5child}},
\code{\link{brv5grand}},
\code{\link{brv5oth}},
\code{\link{brv5par}},
\code{\link{brv5sib}},
\code{\link{brv5sp}},
\code{\link{brv5spnum}}
}
\concept{GSS Next}
\keyword{variable}
