% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{blkwhite}
\alias{blkwhite}
\title{R agrees right & wrong not black & white}
\source{
General Social Survey https://gss.norc.org
}
\description{
blkwhite
}
\details{
Question Please consider the following statements and tell me whether you agree strongly, agree somewhat, disagree somewhat or disagree strongly with each statement.
B. Right and wrong are not usually a simple matter of black and white; there are many shades of gray.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1082/vshow}{GSS Data Explorer page for \code{blkwhite}}.
}

\seealso{
Other Religion: 
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{gochurch}},
\code{\link{permoral}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
