% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_419.R
\name{bmitzvah}
\alias{bmitzvah}
\title{R had bar or bat mitzvah when r was a child}
\source{
General Social Survey https://gss.norc.org
}
\description{
bmitzvah
}
\details{
Question Did you have a (bar/bat) mitzvah when you were a child?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3828/vshow}{GSS Data Explorer page for \code{bmitzvah}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 1983 \tab 23 \tab 17 \tab 2023 \cr
   2012 \tab 1946 \tab 14 \tab 14 \tab 1974 \cr
   Total \tab 3929 \tab 37 \tab 31 \tab 3997 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{kd1relig}},
\code{\link{kd2relig}},
\code{\link{kd3relig}},
\code{\link{kd4relig}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{synmem}}
}
\concept{Jewish Identity}
\keyword{variable}
