% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{books16}
\alias{books16}
\title{Books in home of r age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
books16
}
\details{
Question About how many books were there around your family’s house when you were 16 years old?
}
\section{Values}{

\itemize{
\item \code{1} none
\item \code{2} 1 or 2
\item \code{3} around 10
\item \code{4} around 20
\item \code{5} around 50
\item \code{6} around 100
\item \code{7} around 200
\item \code{8} around 500
\item \code{9} 1,000 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4342/vshow}{GSS Data Explorer page for \code{books16}}.
}

\concept{ISSP Social Inequality}
\concept{Social Class}
\keyword{variable}
