% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{boredom}
\alias{boredom}
\title{How often is r bored at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
boredom
}
\details{
Question Now some more questions about your working conditions. one code for each item below to show how often it applies to your work. How often...
D. Are you bored at work?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4553/vshow}{GSS Data Explorer page for \code{boredom}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 12 \tab 57 \tab 294 \tab 283 \tab 224 \tab 2 \tab 26 \tab 1537 \cr
   Total \tab 639 \tab 12 \tab 57 \tab 294 \tab 283 \tab 224 \tab 2 \tab 26 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
