% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{bornhome}
\alias{bornhome}
\title{What kind of house r's family lived at birth}
\source{
General Social Survey https://gss.norc.org
}
\description{
bornhome
}
\details{
Question Thinking back to the time when you were born, tell me which box shows the kind of house your family lived in at your work: A, B, or C?  We just want you to choose one box.
}
\section{Values}{

\itemize{
\item \code{1} box a
\item \code{2} box b
\item \code{3} box c
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3025/vshow}{GSS Data Explorer page for \code{bornhome}}.
}

\seealso{
Other Immigration: 
\code{\link{immstats}},
\code{\link{natborn}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}}
}
\concept{Dwelling (R's)}
\concept{Immigration}
\keyword{variable}
