% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{bornhome}
\alias{bornhome}
\title{What kind of house r's family lived at birth}
\source{
General Social Survey https://gss.norc.org
}
\description{
bornhome
}
\details{
Question Thinking back to the time when you were born, tell me which box shows the kind of house your family lived in at your work: A, B, or C?  We just want you to choose one box.
}
\section{Values}{

\itemize{
\item \verb{[1]} box a
\item \verb{[2]} box b
\item \verb{[3]} box c
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3025/vshow}{GSS Data Explorer page for \code{bornhome}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab box a \tab box b \tab box c \tab other \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2539 \tab 51 \tab 180 \tab 16 \tab 3 \tab 16 \tab 7 \tab 2812 \cr
   Total \tab 2539 \tab 51 \tab 180 \tab 16 \tab 3 \tab 16 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Immigration: 
\code{\link{immstats}},
\code{\link{natborn}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}}
}
\concept{Immigration}
\keyword{variable}
