% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_455.R
\name{borrow1}
\alias{borrow1}
\title{Borrow money #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
borrow1
}
\details{
Question Suppose you needed to borrow a large sum of money.
A. Who would you turn to first for help?
B. Who would you turn to second?
PLEASE CIRCLE ONLY ONE FIRST CHOICE NUMBER AND ONE SECOND CHOICE NUMBER.
A. First Choice
}
\section{Values}{

\itemize{
\item \code{1} spouse
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other relative
\item \code{9} closest friend
\item \code{10} other friend
\item \code{11} neighbor
\item \code{12} coworker
\item \code{13} bank
\item \code{14} employer
\item \code{15} govt or soc ser
\item \code{16} other
\item \code{17} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4173/vshow}{GSS Data Explorer page for \code{borrow1}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab coworker \tab bank \tab employer \tab govt or soc ser \tab other \tab no one \tab don't know \tab no answer \tab Total \cr
   1986 \tab 42 \tab 191 \tab 142 \tab 143 \tab 32 \tab 46 \tab 50 \tab 50 \tab 65 \tab 58 \tab 3 \tab 4 \tab 547 \tab 10 \tab 4 \tab 3 \tab 53 \tab 1 \tab 26 \tab 1470 \cr
   Total \tab 42 \tab 191 \tab 142 \tab 143 \tab 32 \tab 46 \tab 50 \tab 50 \tab 65 \tab 58 \tab 3 \tab 4 \tab 547 \tab 10 \tab 4 \tab 3 \tab 53 \tab 1 \tab 26 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow2}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{down1}},
\code{\link{down2}},
\code{\link{sick1}},
\code{\link{sick2}},
\code{\link{upset1}},
\code{\link{upset2}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
