% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_455.R
\name{borrow1}
\alias{borrow1}
\title{Borrow money #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
borrow1
}
\details{
Question Suppose you needed to borrow a large sum of money.
A. Who would you turn to first for help?
B. Who would you turn to second?
PLEASE CIRCLE ONLY ONE FIRST CHOICE NUMBER AND ONE SECOND CHOICE NUMBER.
A. First Choice
}
\section{Values}{

\itemize{
\item \code{1} spouse
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other relative
\item \code{9} closest friend
\item \code{10} other friend
\item \code{11} neighbor
\item \code{12} coworker
\item \code{13} bank
\item \code{14} employer
\item \code{15} govt or soc ser
\item \code{16} other
\item \code{17} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4173/vshow}{GSS Data Explorer page for \code{borrow1}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{bestfrd}},
\code{\link{borrow1a}},
\code{\link{borrow2}},
\code{\link{borrow2a}},
\code{\link{brocall}},
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{change1}},
\code{\link{change2}},
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{comyear}},
\code{\link{cousins}},
\code{\link{cowrkfrd}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{demands}},
\code{\link{down1}},
\code{\link{down1a}},
\code{\link{down2}},
\code{\link{down2a}},
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{godparts}},
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{grpchrty}},
\code{\link{grpchurh}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{hapunhap}},
\code{\link{helpfrds}},
\code{\link{helphwrk}},
\code{\link{helpjob}},
\code{\link{inlawnum}},
\code{\link{kidcall}},
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidpars}},
\code{\link{kidvisit}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{localgvt}},
\code{\link{macall}},
\code{\link{macall1}},
\code{\link{malive}},
\code{\link{marcohab}},
\code{\link{matime}},
\code{\link{matime1}},
\code{\link{mavisit}},
\code{\link{mavisit1}},
\code{\link{neifrd}},
\code{\link{niecenep}},
\code{\link{othfrd}},
\code{\link{pacall}},
\code{\link{pacall1}},
\code{\link{palive}},
\code{\link{parslaw}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{pavisit1}},
\code{\link{peocntct}},
\code{\link{posslq}},
\code{\link{posslqy}},
\code{\link{relcall}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{sibcall}},
\code{\link{sibinlaw}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}},
\code{\link{sick1}},
\code{\link{sick1a}},
\code{\link{sick2}},
\code{\link{sick2a}},
\code{\link{siscall}},
\code{\link{sisnum}},
\code{\link{sistime}},
\code{\link{sisvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}},
\code{\link{talkedto}},
\code{\link{trustpeo}},
\code{\link{unaunum}},
\code{\link{uncaunts}},
\code{\link{upset1}},
\code{\link{upset2}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
