% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_459.R
\name{borrow1a}
\alias{borrow1a}
\title{Who does r turn to first when large money needed}
\source{
General Social Survey https://gss.norc.org
}
\description{
borrow1a
}
\details{
Question Now, suppose you needed to borrow a large sum of money.
A. Who would you turn to first for help?
}
\section{Values}{

\itemize{
\item \code{1} husband, wife, partner
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other blood relative
\item \code{9} other in-law relative
\item \code{10} god-parent
\item \code{11} close friend
\item \code{12} neighbor
\item \code{13} someone you work with
\item \code{14} employer
\item \code{15} government or social services agency
\item \code{16} bank or credit union
\item \code{17} private money lender
\item \code{18} someone else
\item \code{19} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4216/vshow}{GSS Data Explorer page for \code{borrow1a}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab husband, wife, partner \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other blood relative \tab other in-law relative \tab close friend \tab someone you work with \tab employer \tab government or social services agency \tab bank or credit union \tab private money lender \tab someone else \tab no one \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 144 \tab 186 \tab 164 \tab 32 \tab 33 \tab 56 \tab 64 \tab 36 \tab 36 \tab 61 \tab 8 \tab 2 \tab 6 \tab 88 \tab 5 \tab 75 \tab 134 \tab 6 \tab 13 \tab 2765 \cr
   Total \tab 1616 \tab 144 \tab 186 \tab 164 \tab 32 \tab 33 \tab 56 \tab 64 \tab 36 \tab 36 \tab 61 \tab 8 \tab 2 \tab 6 \tab 88 \tab 5 \tab 75 \tab 134 \tab 6 \tab 13 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow2a}},
\code{\link{down1a}},
\code{\link{down2a}},
\code{\link{grpchrty}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{helphwrk}},
\code{\link{sick1a}},
\code{\link{sick2a}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
