% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_459.R
\name{borrow2a}
\alias{borrow2a}
\title{Who does r turn to second when large money needed}
\source{
General Social Survey https://gss.norc.org
}
\description{
borrow2a
}
\details{
Question Now, suppose you needed to borrow a large sum of money.
B. And who would you turn to second if you needed to borrow a large sum of money?
}
\section{Values}{

\itemize{
\item \code{1} husband, wife, partner
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other blood relative
\item \code{9} other in-law relative
\item \code{10} god-parent
\item \code{11} close friend
\item \code{12} neighbor
\item \code{13} someone you work with
\item \code{14} employer
\item \code{15} government or social services agency
\item \code{16} bank or credit union
\item \code{17} private money lender
\item \code{18} someone else
\item \code{19} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4217/vshow}{GSS Data Explorer page for \code{borrow2a}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1a}},
\code{\link{down1a}},
\code{\link{down2a}},
\code{\link{grpchrty}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{helphwrk}},
\code{\link{sick1a}},
\code{\link{sick2a}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
