% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{borrowed}
\alias{borrowed}
\title{Have you borrowed $100+ from another person}
\source{
General Social Survey https://gss.norc.org
}
\description{
borrowed
}
\details{
Question Have you borrowed $100 or more from another person (not a bank, business, or other organization) at any time during the past 12 months?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1849/vshow}{GSS Data Explorer page for \code{borrowed}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2168 \tab 154 \tab 578 \tab 4 \tab 2904 \cr
   Total \tab 2168 \tab 154 \tab 578 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{fixer1}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{relfixer}},
\code{\link{satfixes}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
