% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{borrowed}
\alias{borrowed}
\title{Have you borrowed $100+ from another person}
\source{
General Social Survey https://gss.norc.org
}
\description{
borrowed
}
\details{
Question Have you borrowed $100 or more from another person (not a bank, business, or other organization) at any time during the past 12 months?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1849/vshow}{GSS Data Explorer page for \code{borrowed}}.
}

\seealso{
Other Markets: 
\code{\link{fixer1}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{relfixer}},
\code{\link{satfixes}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
