% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{bossemps}
\alias{bossemps}
\title{Quality of mgmnt-employee rels: r's firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
bossemps
}
\details{
Question In general how would you describe relations in your work place…
A. between management and employees?
}
\section{Values}{

\itemize{
\item \code{1} very good
\item \code{2} quite good
\item \code{3} neither good nor bad
\item \code{4} quite bad
\item \code{5} very bad
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998, 2012 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4559/vshow}{GSS Data Explorer page for \code{bossemps}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very good \tab quite good \tab neither good nor bad \tab quite bad \tab very bad \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 191 \tab 392 \tab 210 \tab 46 \tab 12 \tab 21 \tab 26 \tab 1537 \cr
   1998 \tab 1967 \tab 222 \tab 345 \tab 178 \tab 42 \tab 18 \tab 20 \tab 40 \tab 2832 \cr
   2006 \tab 3492 \tab 334 \tab 360 \tab 232 \tab 61 \tab 24 \tab 4 \tab 3 \tab 4510 \cr
   2012 \tab 795 \tab 350 \tab 446 \tab 254 \tab 56 \tab 13 \tab 24 \tab 36 \tab 1974 \cr
   2016 \tab 1927 \tab 310 \tab 344 \tab 205 \tab 55 \tab 10 \tab 8 \tab 8 \tab 2867 \cr
   Total \tab 8820 \tab 1407 \tab 1887 \tab 1079 \tab 260 \tab 77 \tab 77 \tab 113 \tab 13720 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
