% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{bosswrks}
\alias{bosswrks}
\title{Management & workers always in conflict}
\source{
General Social Survey https://gss.norc.org
}
\description{
bosswrks
}
\details{
Question How much do you agree or disagree with these two statements?
A. There will always be conflict between management and workers because they are really on opposite sides
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4505/vshow}{GSS Data Explorer page for \code{bosswrks}}.
}

\seealso{
Other Work Orientation: 
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{hiinc}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}},
\code{\link{strngun}}
}
\concept{Work Orientation}
\keyword{variable}
