% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{bribe}
\alias{bribe}
\title{Does r see public official asking for a bribe in last five years}
\source{
General Social Survey https://gss.norc.org
}
\description{
bribe
}
\details{
Question In the last five years, how often have you or a member of your immediate family come across a public official who hinted they wanted, or asked for, a bribe or favour in return for a service?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} seldom
\item \code{3} occasionally
\item \code{4} quite often
\item \code{5} very often
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4125/vshow}{GSS Data Explorer page for \code{bribe}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{corrupt1}},
\code{\link{corrupt2}},
\code{\link{knowpols}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
