% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{brkdas1}
\alias{brkdas1}
\title{1st mention, what nervous breakdown is, asserted}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdas1
}
\details{
Question As far as you know, what is a nervous breakdown? (PROBES: How would you describe it?  What is it like?  What happens to a person who has one? How does he act?)
\enumerate{
\item R's asserted references for what a nervous breakdown is -- First mention.
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2049/vshow}{GSS Data Explorer page for \code{brkdas1}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 21 \tab 22 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1996 \tab 2168 \tab 22 \tab 27 \tab 3 \tab 17 \tab 35 \tab 49 \tab 1 \tab 1 \tab 3 \tab 1 \tab 9 \tab 9 \tab 3 \tab 23 \tab 81 \tab 348 \tab 5 \tab 20 \tab 40 \tab 7 \tab 32 \tab 2904 \cr
   Total \tab 2168 \tab 22 \tab 27 \tab 3 \tab 17 \tab 35 \tab 49 \tab 1 \tab 1 \tab 3 \tab 1 \tab 9 \tab 9 \tab 3 \tab 23 \tab 81 \tab 348 \tab 5 \tab 20 \tab 40 \tab 7 \tab 32 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
