% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{brkdas2}
\alias{brkdas2}
\title{2nd mention, what nervous breakdown is, asserted}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdas2
}
\details{
Question As far as you know, what is a nervous breakdown? (PROBES: How would you describe it?  What is it like?  What happens to a person who has one? How does he act?)
2. R's asserted references for what a nervous breakdown is -- Second mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2050/vshow}{GSS Data Explorer page for \code{brkdas2}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 16 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1996 \tab 2633 \tab 9 \tab 1 \tab 16 \tab 3 \tab 3 \tab 16 \tab 2 \tab 1 \tab 10 \tab 1 \tab 2 \tab 1 \tab 12 \tab 46 \tab 65 \tab 4 \tab 9 \tab 28 \tab 10 \tab 32 \tab 2904 \cr
   Total \tab 2633 \tab 9 \tab 1 \tab 16 \tab 3 \tab 3 \tab 16 \tab 2 \tab 1 \tab 10 \tab 1 \tab 2 \tab 1 \tab 12 \tab 46 \tab 65 \tab 4 \tab 9 \tab 28 \tab 10 \tab 32 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas1}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}}
}
\concept{Mental Health}
\keyword{variable}
