% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{brkddo1}
\alias{brkddo1}
\title{1st mention what r did when felt ner. break}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkddo1
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
B. (IF EVER ANTICIPATED A NERVOUS BREAKDOWN) What did you do about it? (PROBE: Anything
else? until R says no.)  Which of these things did you do first; second; third?
(INTERVIEWER: Sequence list; read back if needed.)
B1. What R did about the nervous breakdown - First mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2070/vshow}{GSS Data Explorer page for \code{brkddo1}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 11 \tab 21 \tab 22 \tab 23 \tab 24 \tab 31 \tab 41 \tab 51 \tab 61 \tab 62 \tab 64 \tab 65 \tab 69 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 81 \tab 82 \tab 83 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2517 \tab 2 \tab 1 \tab 11 \tab 5 \tab 3 \tab 49 \tab 4 \tab 18 \tab 8 \tab 3 \tab 13 \tab 1 \tab 11 \tab 10 \tab 25 \tab 8 \tab 18 \tab 9 \tab 66 \tab 90 \tab 5 \tab 1 \tab 26 \tab 2904 \cr
   Total \tab 2517 \tab 2 \tab 1 \tab 11 \tab 5 \tab 3 \tab 49 \tab 4 \tab 18 \tab 8 \tab 3 \tab 13 \tab 1 \tab 11 \tab 10 \tab 25 \tab 8 \tab 18 \tab 9 \tab 66 \tab 90 \tab 5 \tab 1 \tab 26 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{evbrkdwn}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}}
}
\concept{Mental Health}
\keyword{variable}
