% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{brkddo3}
\alias{brkddo3}
\title{3rd mention what r did when felt ner. break}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkddo3
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
B3. What R did about the nervous breakdown -- Third mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2072/vshow}{GSS Data Explorer page for \code{brkddo3}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 11 \tab 22 \tab 23 \tab 24 \tab 51 \tab 61 \tab 62 \tab 64 \tab 65 \tab 66 \tab 69 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 81 \tab 82 \tab 83 \tab no answer \tab Total \cr
   1996 \tab 2767 \tab 1 \tab 8 \tab 2 \tab 1 \tab 6 \tab 10 \tab 3 \tab 5 \tab 2 \tab 1 \tab 5 \tab 9 \tab 9 \tab 2 \tab 1 \tab 4 \tab 21 \tab 18 \tab 3 \tab 26 \tab 2904 \cr
   Total \tab 2767 \tab 1 \tab 8 \tab 2 \tab 1 \tab 6 \tab 10 \tab 3 \tab 5 \tab 2 \tab 1 \tab 5 \tab 9 \tab 9 \tab 2 \tab 1 \tab 4 \tab 21 \tab 18 \tab 3 \tab 26 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdtime}},
\code{\link{evmhp}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
