% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{brkddur}
\alias{brkddur}
\title{Duration, severity of nervous breakdown feelings}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkddur
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
E. The duration and severity of R''s breakdown.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2077/vshow}{GSS Data Explorer page for \code{brkddur}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab no answer \tab Total \cr
   1996 \tab 2517 \tab 87 \tab 1 \tab 18 \tab 16 \tab 20 \tab 5 \tab 51 \tab 13 \tab 176 \tab 2904 \cr
   Total \tab 2517 \tab 87 \tab 1 \tab 18 \tab 16 \tab 20 \tab 5 \tab 51 \tab 13 \tab 176 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdtime}},
\code{\link{evmhp}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
