% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{brkdoth}
\alias{brkdoth}
\title{What nervous breakdown is, other codes}
\source{
General Social Survey https://gss.norc.org
}
\description{
brkdoth
}
\details{
Question As far as you know, what is a nervous breakdown? (PROBES: How would you describe it?  What is it like?  What happens to a person who has one? How does he act?)
\enumerate{
\item Other mentions about the nature of a nervous breakdown.
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2057/vshow}{GSS Data Explorer page for \code{brkdoth}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 1002 \tab 1003 \tab 1004 \tab 1008 \tab 1010 \tab 1104 \tab 1109 \tab 1203 \tab 1207 \tab 1209 \tab 1210 \tab 1211 \tab 1214 \tab no answer \tab Total \cr
   1996 \tab 2741 \tab 1 \tab 2 \tab 2 \tab 3 \tab 7 \tab 20 \tab 65 \tab 1 \tab 2 \tab 11 \tab 1 \tab 26 \tab 9 \tab 13 \tab 2904 \cr
   Total \tab 2741 \tab 1 \tab 2 \tab 2 \tab 3 \tab 7 \tab 20 \tab 65 \tab 1 \tab 2 \tab 11 \tab 1 \tab 26 \tab 9 \tab 13 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{knwmhosp}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}}
}
\concept{Mental Health}
\keyword{variable}
