% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{brneffrt}
\alias{brneffrt}
\title{Possible to become american if born elsewhere}
\source{
General Social Survey https://gss.norc.org
}
\description{
brneffrt
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} i definitely agree with statement a
\item \code{2} i agree more with statement a than with statement b
\item \code{3} i agree more with statement b than with statement a
\item \code{4} i definitely agree with statement b
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
