% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{brneffrt}
\alias{brneffrt}
\title{Possible to become american if born elsewhere}
\source{
General Social Survey https://gss.norc.org
}
\description{
brneffrt
}
\details{
Question Some people say that it is possible to become truly American. if a person makes an effort. Others say a person has to be born American to be truly American. What is your position?
}
\section{Values}{

\itemize{
\item \code{1} i definitely agree with statement a
\item \code{2} i agree more with statement a than with statement b
\item \code{3} i agree more with statement b than with statement a
\item \code{4} i definitely agree with statement b
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8172/vshow}{GSS Data Explorer page for \code{brneffrt}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{conscrp}},
\code{\link{immcrime1}},
\code{\link{immideas1}},
\code{\link{immjobs1}},
\code{\link{immwlfare}}
}
\concept{ISSP National Identity}
\keyword{variable}
