% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{brneffrt}
\alias{brneffrt}
\title{Possible to become american if born elsewhere}
\source{
General Social Survey https://gss.norc.org
}
\description{
brneffrt
}
\details{
Question Some people say that it is possible to become truly American. if a person makes an effort. Others say a person has to be born American to be truly American. What is your position?
}
\section{Values}{

\itemize{
\item \code{1} i definitely agree with statement a
\item \code{2} i agree more with statement a than with statement b
\item \code{3} i agree more with statement b than with statement a
\item \code{4} i definitely agree with statement b
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8172/vshow}{GSS Data Explorer page for \code{brneffrt}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amproud1}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{belikeus}},
\code{\link{biasnews}},
\code{\link{bizmoney}},
\code{\link{citizen}},
\code{\link{citworld}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{decsorgs}},
\code{\link{ethclose}},
\code{\link{experts}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{freetrde}},
\code{\link{fulldem}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immassim}},
\code{\link{immcosts}},
\code{\link{immcrime1}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{immideas}},
\code{\link{immideas1}},
\code{\link{immimp}},
\code{\link{immjobs1}},
\code{\link{immrghts}},
\code{\link{immwlfare}},
\code{\link{imports}},
\code{\link{incgap1}},
\code{\link{internet}},
\code{\link{intlincs}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{lessprd}},
\code{\link{letin1}},
\code{\link{letin1a}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{refugees}},
\code{\link{res161}},
\code{\link{shortcom}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
