% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{brnotax}
\alias{brnotax}
\title{Do private companies avoid taxes}
\source{
General Social Survey https://gss.norc.org
}
\description{
brnotax
}
\details{
Question (In general, how often do you think that major private companies in America do the following?
Try to avoid paying their taxes?  (Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6314/vshow}{GSS Data Explorer page for \code{brnotax}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 372 \tab 339 \tab 466 \tab 153 \tab 45 \tab 15 \tab 2867 \cr
   Total \tab 1477 \tab 372 \tab 339 \tab 466 \tab 153 \tab 45 \tab 15 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
