% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{broadband}
\alias{broadband}
\title{Household has access to internet using broadband internet service}
\source{
General Social Survey https://gss.norc.org
}
\description{
broadband
}
\details{
Question Do you or any member of this household have access to the Internet using a broadband (high speed) Internet service such as cable, fiber optic, or DSL service installed in this household?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8195/vshow}{GSS Data Explorer page for \code{broadband}}.
}

\concept{Core}
\keyword{variable}
