% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{brocall}
\alias{brocall}
\title{Contact with brother}
\source{
General Social Survey https://gss.norc.org
}
\description{
brocall
}
\details{
Question How many brothers age 18 or older do you have? (We mean brothers who are still alive; please include step-brothers, half-brothers, and adopted brothers.)
C. And how often do you have any other contact with your brother  besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} several x week
\item \code{3} once a week
\item \code{4} once a month
\item \code{5} several x year
\item \code{6} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4142/vshow}{GSS Data Explorer page for \code{brocall}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{posslq}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
