% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{brocall}
\alias{brocall}
\title{Contact with brother}
\source{
General Social Survey https://gss.norc.org
}
\description{
brocall
}
\details{
Question How many brothers age 18 or older do you have? (We mean brothers who are still alive; please include step-brothers, half-brothers, and adopted brothers.)
C. And how often do you have any other contact with your brother  besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several x week
\item \verb{[3]} once a week
\item \verb{[4]} once a month
\item \verb{[5]} several x year
\item \verb{[6]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4142/vshow}{GSS Data Explorer page for \code{brocall}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab don't know \tab no answer \tab Total \cr
   1986 \tab 462 \tab 42 \tab 77 \tab 175 \tab 254 \tab 227 \tab 222 \tab 1 \tab 10 \tab 1470 \cr
   Total \tab 462 \tab 42 \tab 77 \tab 175 \tab 254 \tab 227 \tab 222 \tab 1 \tab 10 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{posslq}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
