% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{browser3}
\alias{browser3}
\title{R specify name of web browser}
\source{
General Social Survey https://gss.norc.org
}
\description{
browser3
}
\details{
Question D. Specify name of web browser
}
\section{Values}{

\itemize{
\item \code{1} explorer
\item \code{2} netscape
\item \code{3} excite
\item \code{4} msn
\item \code{5} aol
\item \code{6} google
\item \code{7} yahoo
\item \code{8} earthlink
\item \code{9} lycos
\item \code{10} compuserve
\item \code{11} bonzi
\item \code{12} goto
\item \code{13} altavista
\item \code{14} hotbot
\item \code{15} siap
\item \code{16} mindspring
\item \code{17} dogpile
\item \code{18} prodigy
\item \code{19} ask jeeves
\item \code{20} dsl
\item \code{21} netrax
\item \code{22} webcorner
\item \code{23} mozilla
\item \code{24} everyonesinternet
\item \code{25} mamma
\item \code{26} dogpatch
\item \code{27} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2633/vshow}{GSS Data Explorer page for \code{browser3}}.
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{ezines}},
\code{\link{mp3}},
\code{\link{newsgrps}},
\code{\link{prefsets}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}}
}
\concept{Information Society}
\keyword{variable}
