% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{browser3}
\alias{browser3}
\title{R specify name of web browser}
\source{
General Social Survey https://gss.norc.org
}
\description{
browser3
}
\details{
Question D. Specify name of web browser
}
\section{Values}{

\itemize{
\item \verb{[1]} explorer
\item \verb{[2]} netscape
\item \verb{[3]} excite
\item \verb{[4]} msn
\item \verb{[5]} aol
\item \verb{[6]} google
\item \verb{[7]} yahoo
\item \verb{[8]} earthlink
\item \verb{[9]} lycos
\item \verb{[10]} compuserve
\item \verb{[11]} bonzi
\item \verb{[12]} goto
\item \verb{[13]} altavista
\item \verb{[14]} hotbot
\item \verb{[15]} siap
\item \verb{[16]} mindspring
\item \verb{[17]} dogpile
\item \verb{[18]} prodigy
\item \verb{[19]} ask jeeves
\item \verb{[20]} dsl
\item \verb{[21]} netrax
\item \verb{[22]} webcorner
\item \verb{[23]} mozilla
\item \verb{[24]} everyonesinternet
\item \verb{[25]} mamma
\item \verb{[26]} dogpatch
\item \verb{[27]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2633/vshow}{GSS Data Explorer page for \code{browser3}}.

Counts by year:\tabular{llllllllllllllllllllllllllllll}{
   year \tab iap \tab explorer \tab netscape \tab excite \tab msn \tab aol \tab google \tab yahoo \tab earthlink \tab lycos \tab compuserve \tab bonzi \tab goto \tab altavista \tab hotbot \tab siap \tab mindspring \tab dogpile \tab prodigy \tab ask jeeves \tab dsl \tab netrax \tab webcorner \tab everyonesinternet \tab mamma \tab other \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2338 \tab 119 \tab 66 \tab 2 \tab 16 \tab 27 \tab 21 \tab 93 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 2 \tab 47 \tab 2765 \cr
   Total \tab 2338 \tab 119 \tab 66 \tab 2 \tab 16 \tab 27 \tab 21 \tab 93 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 2 \tab 47 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{ezines}},
\code{\link{mp3}},
\code{\link{newsgrps}},
\code{\link{prefsets}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}}
}
\concept{Information Society}
\keyword{variable}
