% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_659.R
\name{brv16}
\alias{brv16}
\title{Death of someone clsoe before age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16
}
\details{
Question Before the time you were 16, did you experience the death of someone close to you?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7921/vshow}{GSS Data Explorer page for \code{brv16}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 768 \tab 465 \tab 2 \tab 3544 \cr
   Total \tab 2309 \tab 768 \tab 465 \tab 2 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16child}},
\code{\link{brv16grand}},
\code{\link{brv16oth}},
\code{\link{brv16par}},
\code{\link{brv16sib}},
\code{\link{brv16sp}},
\code{\link{brv5cowknum}},
\code{\link{brv5frndnum}},
\code{\link{brv5othnum}}
}
\concept{GSS Next}
\keyword{variable}
