% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_660.R
\name{brv16gmanum}
\alias{brv16gmanum}
\title{Number of grandmother deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv16gmanum
}
\details{
Question NUMBER OF GRANDMOTHER DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{4} one or more, number unknown
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7934/vshow}{GSS Data Explorer page for \code{brv16gmanum}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 25 \tab 3 \tab 4 \tab skipped on web \tab Total \cr
   2022 \tab 3009 \tab 233 \tab 159 \tab 34 \tab 1 \tab 2 \tab 1 \tab 105 \tab 3544 \cr
   Total \tab 3009 \tab 233 \tab 159 \tab 34 \tab 1 \tab 2 \tab 1 \tab 105 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16dadnum}},
\code{\link{brv16daunum}},
\code{\link{brv16filnum}},
\code{\link{brv16gpanum}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16partnum}},
\code{\link{brv16sonnum}},
\code{\link{brv16spnum}}
}
\concept{GSS Next}
\keyword{variable}
