% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_657.R
\name{brv5}
\alias{brv5}
\title{Death of someone close in past 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5
}
\details{
Question These next questions are about your experiences losing people close to you. Thinking back five years, to the period from January 1, 2017 through today, did you experience the death of someone close to you?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7896/vshow}{GSS Data Explorer page for \code{brv5}}.
}

\seealso{
Other GSS Next: 
\code{\link{blkdsrv}},
\code{\link{blktry}},
\code{\link{brv5child}},
\code{\link{brv5grand}},
\code{\link{brv5oth}},
\code{\link{brv5par}},
\code{\link{brv5sib}},
\code{\link{brv5sp}},
\code{\link{brv5spnum}}
}
\concept{GSS Next}
\keyword{variable}
