% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_659.R
\name{brv5bronum}
\alias{brv5bronum}
\title{Number of brother deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5bronum
}
\details{
Question NUMBER OF BROTHER DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7914/vshow}{GSS Data Explorer page for \code{brv5bronum}}.
}

\seealso{
Other GSS Next: 
\code{\link{brv16}},
\code{\link{brv16par}},
\code{\link{brv16sp}},
\code{\link{brv5cowknum}},
\code{\link{brv5cuznum}},
\code{\link{brv5frndnum}},
\code{\link{brv5othnum}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}}
}
\concept{GSS Next}
\keyword{variable}
