% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_659.R
\name{brv5cowknum}
\alias{brv5cowknum}
\title{Number of coworker deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5cowknum
}
\details{
Question NUMBER OF COWORKER DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7919/vshow}{GSS Data Explorer page for \code{brv5cowknum}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 20 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab skipped on web \tab Total \cr
   2022 \tab 3070 \tab 108 \tab 73 \tab 22 \tab 1 \tab 15 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 245 \tab 3544 \cr
   Total \tab 3070 \tab 108 \tab 73 \tab 22 \tab 1 \tab 15 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 245 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv16}},
\code{\link{brv16child}},
\code{\link{brv16grand}},
\code{\link{brv16oth}},
\code{\link{brv16par}},
\code{\link{brv16sib}},
\code{\link{brv16sp}},
\code{\link{brv5frndnum}},
\code{\link{brv5othnum}}
}
\concept{GSS Next}
\keyword{variable}
