% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_658.R
\name{brv5cuznum}
\alias{brv5cuznum}
\title{Number of cousin deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5cuznum
}
\details{
Question NUMBER OF COUSIN DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7917/vshow}{GSS Data Explorer page for \code{brv5cuznum}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 7 \tab 8 \tab skipped on web \tab Total \cr
   2022 \tab 3070 \tab 103 \tab 99 \tab 1 \tab 46 \tab 11 \tab 6 \tab 4 \tab 1 \tab 1 \tab 202 \tab 3544 \cr
   Total \tab 3070 \tab 103 \tab 99 \tab 1 \tab 46 \tab 11 \tab 6 \tab 4 \tab 1 \tab 1 \tab 202 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv5bronum}},
\code{\link{brv5chinum}},
\code{\link{brv5daunum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5milnum}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}},
\code{\link{brv5sonnum}}
}
\concept{GSS Next}
\keyword{variable}
