% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_657.R
\name{brv5dadnum}
\alias{brv5dadnum}
\title{Number of father deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5dadnum
}
\details{
Question NUMBER OF FATHER DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{4} one or more, number unknown
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7905/vshow}{GSS Data Explorer page for \code{brv5dadnum}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 2 \tab 5 \tab skipped on web \tab Total \cr
   2022 \tab 3238 \tab 88 \tab 125 \tab 1 \tab 5 \tab 1 \tab 86 \tab 3544 \cr
   Total \tab 3238 \tab 88 \tab 125 \tab 1 \tab 5 \tab 1 \tab 86 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv5child}},
\code{\link{brv5filnum}},
\code{\link{brv5grand}},
\code{\link{brv5momnum}},
\code{\link{brv5oth}},
\code{\link{brv5par}},
\code{\link{brv5partnum}},
\code{\link{brv5sib}},
\code{\link{brv5spnum}}
}
\concept{GSS Next}
\keyword{variable}
