% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_658.R
\name{brv5filnum}
\alias{brv5filnum}
\title{Number of father in law deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5filnum
}
\details{
Question NUMBER OF FATHER IN LAW DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7907/vshow}{GSS Data Explorer page for \code{brv5filnum}}.
}

\seealso{
Other GSS Next: 
\code{\link{brv5chinum}},
\code{\link{brv5dadnum}},
\code{\link{brv5daunum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5milnum}},
\code{\link{brv5momnum}},
\code{\link{brv5partnum}},
\code{\link{brv5sonnum}}
}
\concept{GSS Next}
\keyword{variable}
