% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_658.R
\name{brv5milnum}
\alias{brv5milnum}
\title{Number of mother in law deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5milnum
}
\details{
Question NUMBER OF MOTHER IN LAW DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7908/vshow}{GSS Data Explorer page for \code{brv5milnum}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab skipped on web \tab Total \cr
   2022 \tab 3239 \tab 121 \tab 46 \tab 2 \tab 136 \tab 3544 \cr
   Total \tab 3239 \tab 121 \tab 46 \tab 2 \tab 136 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv5bronum}},
\code{\link{brv5chinum}},
\code{\link{brv5cuznum}},
\code{\link{brv5daunum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}},
\code{\link{brv5sonnum}}
}
\concept{GSS Next}
\keyword{variable}
