% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_657.R
\name{brv5partnum}
\alias{brv5partnum}
\title{Number of partner deaths}
\source{
General Social Survey https://gss.norc.org
}
\description{
brv5partnum
}
\details{
Question NUMBER OF PARTNER DEATHS

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7904/vshow}{GSS Data Explorer page for \code{brv5partnum}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 0 \tab 1 \tab skipped on web \tab Total \cr
   2022 \tab 3476 \tab 10 \tab 20 \tab 38 \tab 3544 \cr
   Total \tab 3476 \tab 10 \tab 20 \tab 38 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{brv5child}},
\code{\link{brv5dadnum}},
\code{\link{brv5filnum}},
\code{\link{brv5grand}},
\code{\link{brv5momnum}},
\code{\link{brv5oth}},
\code{\link{brv5par}},
\code{\link{brv5sib}},
\code{\link{brv5spnum}}
}
\concept{GSS Next}
\keyword{variable}
