% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{bstcall}
\alias{bstcall}
\title{How often does r call best friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
bstcall
}
\details{
Question And how often do you have any other contact with this friend besides visiting, either by telephone, letter, fax or e-mail?
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} at least several times a week
\item \code{3} at least once a week
\item \code{4} at least once a month
\item \code{5} several times a year
\item \code{6} less often
\item \code{7} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4206/vshow}{GSS Data Explorer page for \code{bstcall}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab daily \tab at least several times a week \tab at least once a week \tab at least once a month \tab several times a year \tab less often \tab never \tab no answer \tab Total \cr
   2002 \tab 1701 \tab 192 \tab 282 \tab 241 \tab 169 \tab 80 \tab 50 \tab 43 \tab 7 \tab 2765 \cr
   Total \tab 1701 \tab 192 \tab 282 \tab 241 \tab 169 \tab 80 \tab 50 \tab 43 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
