% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_603.R
\name{bthgrp2f}
\alias{bthgrp2f}
\title{What are the groups you both belong to?}
\source{
General Social Survey https://gss.norc.org
}
\description{
bthgrp2f
}
\details{
Question What are the groups you both belong to? Circle all that apply.
}
\section{Values}{

\itemize{
\item \verb{[1]} fraternal groups
\item \verb{[2]} service clubs
\item \verb{[3]} veterans groups
\item \verb{[4]} political clubs
\item \verb{[5]} labor unions
\item \verb{[6]} sports groups
\item \verb{[7]} youth groups
\item \verb{[8]} school service groups
\item \verb{[9]} hobby or garden clubs
\item \verb{[10]} school fraternities or sororities
\item \verb{[11]} nationality groups
\item \verb{[12]} farm organizations
\item \verb{[13]} literary art discussion or study grps
\item \verb{[14]} professional or academic societies
\item \verb{[15]} church-affiliated groups
\item \verb{[16]} other groups
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5526/vshow}{GSS Data Explorer page for \code{bthgrp2f}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab church-affiliated groups \tab no answer \tab Total \cr
   2004 \tab 2770 \tab 1 \tab 41 \tab 2812 \cr
   Total \tab 2770 \tab 1 \tab 41 \tab 2812 \cr
}
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
