% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{buddhsts}
\alias{buddhsts}
\title{What is your personal attitude towards members of the following religious group}
\source{
General Social Survey https://gss.norc.org
}
\description{
buddhsts
}
\details{
Question What is your personal attitude towards members of the following religious groups? Buddhists
}
\section{Values}{

\itemize{
\item \code{1} very positive
\item \code{2} somewhat positive
\item \code{3} neither positive nor negative
\item \code{4} somewhat negative
\item \code{5} very negative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7289/vshow}{GSS Data Explorer page for \code{buddhsts}}.
}

\keyword{variable}
