% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{buddhsts}
\alias{buddhsts}
\title{What is your personal attitude towards members of the following religious group}
\source{
General Social Survey https://gss.norc.org
}
\description{
buddhsts
}
\details{
Question What is your personal attitude towards members of the following religious groups? Buddhists
}
\section{Values}{

\itemize{
\item \code{1} very positive
\item \code{2} somewhat positive
\item \code{3} neither positive nor negative
\item \code{4} somewhat negative
\item \code{5} very negative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7289/vshow}{GSS Data Explorer page for \code{buddhsts}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very positive \tab somewhat positive \tab neither positive nor negative \tab somewhat negative \tab very negative \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1173 \tab 199 \tab 299 \tab 496 \tab 51 \tab 58 \tab 56 \tab 16 \tab 2348 \cr
   Total \tab 1173 \tab 199 \tab 299 \tab 496 \tab 51 \tab 58 \tab 56 \tab 16 \tab 2348 \cr
}
}

\keyword{variable}
