% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{bullied}
\alias{bullied}
\title{Situation caused by having been bullied}
\source{
General Social Survey https://gss.norc.org
}
\description{
bullied
}
\details{
Question In your opinion, how likely is it that (VIGNAME)’s situation might be caused by having been bullied?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8131/vshow}{GSS Data Explorer page for \code{bullied}}.
}

\seealso{
Other Mental Health: 
\code{\link{bullpeer}},
\code{\link{imprvown1}},
\code{\link{morextrm}},
\code{\link{mustdoc1}},
\code{\link{musthosp1}},
\code{\link{mustmed1}},
\code{\link{physill1}},
\code{\link{smedia}}
}
\concept{Mental Health}
\keyword{variable}
