% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{busdecid}
\alias{busdecid}
\title{Govt should let biz decide how to protect envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
busdecid
}
\details{
Question B. And which one of the following closest to your views?  Government should let businesses decide for themselves how to protect the environment, even if it means they don't always do the right thing, or government should pass laws to make businesses protect the environment, even if it interferes with business' right to make their own decisions.
}
\section{Values}{

\itemize{
\item \code{1} gov. let busnesses decide
\item \code{2} gov. make laws
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4767/vshow}{GSS Data Explorer page for \code{busdecid}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\keyword{variable}
