% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_634.R
\name{buseqinc}
\alias{buseqinc}
\title{Pay difference in companies}
\source{
General Social Survey https://gss.norc.org
}
\description{
buseqinc
}
\details{
Question Think of a score of 1 as meaning that major companies ought to reduce the pay differences between employees with high pay and those with low pay, and a score of 7 meaning that major companies should not concern themselves with reducing pay differences.
}
\section{Values}{

\itemize{
\item \code{1} should do something to reduce pay differences
\item \code{7} should not concern themselves with pay differences
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7546/vshow}{GSS Data Explorer page for \code{buseqinc}}.
}

\keyword{variable}
