% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{busing10}
\alias{busing10}
\title{Favor busing 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
busing10
}
\details{
Question B. Now, thinking about ten years ago, that is in 1972, did you then favor or oppose the busing of (Negro/Black) and white school children from one school district to another?
}
\section{Values}{

\itemize{
\item \verb{[1]} favor
\item \verb{[2]} oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/399/vshow}{GSS Data Explorer page for \code{busing10}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab favor \tab oppose \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 482 \tab 1196 \tab 169 \tab 13 \tab 1860 \cr
   Total \tab 0 \tab 482 \tab 1196 \tab 169 \tab 13 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\keyword{variable}
