% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_442.R
\name{bustax}
\alias{bustax}
\title{Amount of tax on business and industry}
\source{
General Social Survey https://gss.norc.org
}
\description{
bustax
}
\details{
Question Do you consider the amount of tax that business and industry have to pay is too high or too low?
}
\section{Values}{

\itemize{
\item \code{1} much too high
\item \code{2} too high
\item \code{3} about right
\item \code{4} too low
\item \code{5} much too low
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4066/vshow}{GSS Data Explorer page for \code{bustax}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}}

Other Taxes: 
\code{\link{inctax}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}}
}
\concept{Business}
\concept{ISSP Role of Gov't}
\concept{Taxes}
\keyword{variable}
