% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{buyesop}
\alias{buyesop}
\title{Does employee ownership matter when r thinks about buying from a company}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyesop
}
\details{
Question If you were thinking of buying from a company, would you be influenced by whether the company was employee-owned?
PLEASE READ THE DEFINITION OF EMPLOYEE-OWNED: Employee-owned is when a company is owned by all of its employees,
with each employee having a share of ownership.
}
\section{Values}{

\itemize{
\item \verb{[1]} i would be less likely to buy from an employee-owned company
\item \verb{[2]} i would be neither more nor less likely to buy from an employee-owned company
\item \verb{[3]} i would be more likely to buy from an employee-owned company
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7225/vshow}{GSS Data Explorer page for \code{buyesop}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab i would be less likely to buy from an employee-owned company \tab i would be neither more nor less likely to buy from an employee-owned company \tab i would be more likely to buy from an employee-owned company \tab don't know \tab no answer \tab Total \cr
   2018 \tab 900 \tab 121 \tab 753 \tab 517 \tab 26 \tab 31 \tab 2348 \cr
   Total \tab 900 \tab 121 \tab 753 \tab 517 \tab 26 \tab 31 \tab 2348 \cr
}
}

\keyword{variable}
