% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{buyesop}
\alias{buyesop}
\title{Does employee ownership matter when r thinks about buying from a company}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyesop
}
\details{
Question If you were thinking of buying from a company, would you be influenced by whether the company was employee-owned?
PLEASE READ THE DEFINITION OF EMPLOYEE-OWNED: Employee-owned is when a company is owned by all of its employees,
with each employee having a share of ownership.
}
\section{Values}{

\itemize{
\item \code{1} i would be less likely to buy from an employee-owned company
\item \code{2} i would be neither more nor less likely to buy from an employee-owned company
\item \code{3} i would be more likely to buy from an employee-owned company
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7225/vshow}{GSS Data Explorer page for \code{buyesop}}.
}

\keyword{variable}
