% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_197.R
\name{buyfixes}
\alias{buyfixes}
\title{Would you buy home maintenance from a business}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyfixes
}
\details{
Question I'd like you to imagine that you want to buy each of the following items or services. Would you rather buy them from a business or service provider with whom you had some family or social relationship, or from a seller or provider with whom you had no prior personal contact. Please tell me the number that best represents your feeling with respect to each of the purchases listed.
D. Home maintenance (e.g., carpentry, painting)
}
\section{Values}{

\itemize{
\item \verb{[1]} strong preference no contact
\item \verb{[3]} dont care
\item \verb{[5]} strong preference contact
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1860/vshow}{GSS Data Explorer page for \code{buyfixes}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strong preference no contact \tab 2 \tab dont care \tab 4 \tab strong preference contact \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2196 \tab 61 \tab 23 \tab 178 \tab 158 \tab 280 \tab 7 \tab 1 \tab 2904 \cr
   Total \tab 2196 \tab 61 \tab 23 \tab 178 \tab 158 \tab 280 \tab 7 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{sellauto}},
\code{\link{sellhome}},
\code{\link{transoth}}
}
\concept{Markets}
\keyword{variable}
