% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_266.R
\name{buyinf12}
\alias{buyinf12}
\title{R use web for product information in past 12 mo}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyinf12
}
\details{
Question In the past 12 months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/02 have you used the Web to do each of these things:
C. Look for information about a product that you might want to buy?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2417/vshow}{GSS Data Explorer page for \code{buyinf12}}.
}

\seealso{
Other Information Society: 
\code{\link{buyit12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{invest12}},
\code{\link{person30}},
\code{\link{porn30}},
\code{\link{sci30}},
\code{\link{work12}}
}
\concept{Information Society}
\keyword{variable}
