% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_266.R
\name{buyit12}
\alias{buyit12}
\title{R use web to purchase products in past 12 month}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyit12
}
\details{
Question In the past 12 months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/02 have you used the Web to do each of these things:
D. Actually buy something?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2418/vshow}{GSS Data Explorer page for \code{buyit12}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 337 \tab 329 \tab 76 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 623 \tab 366 \tab 94 \tab 1 \tab 2765 \cr
   Total \tab 3756 \tab 960 \tab 695 \tab 170 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{buyinf12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{invest12}},
\code{\link{person30}},
\code{\link{porn30}},
\code{\link{sci30}},
\code{\link{work12}}
}
\concept{Information Society}
\keyword{variable}
