% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_487.R
\name{buythngs}
\alias{buythngs}
\title{Who has final say about buying things for home}
\source{
General Social Survey https://gss.norc.org
}
\description{
buythngs
}
\details{
Question When you and your spouse/partner make decisions about the following, who has the final say?
B. Buying major things for the home
}
\section{Values}{

\itemize{
\item \code{1} mostly me
\item \code{2} mostly my spouse
\item \code{3} sometimes me or sometimes my spouse
\item \code{4} we decide together
\item \code{5} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4442/vshow}{GSS Data Explorer page for \code{buythngs}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famvswrk}},
\code{\link{jobvsfam}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{weekend}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
