% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_390.R
\name{buyvalue}
\alias{buyvalue}
\title{Percent of company stock r bought from own money}
\source{
General Social Survey https://gss.norc.org
}
\description{
buyvalue
}
\details{
Question Approximately what percent of the company stock that you own today did you buy with your own savings or money from your wages?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3551/vshow}{GSS Data Explorer page for \code{buyvalue}}.
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{trdestck}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
