% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{bwfriend}
\alias{bwfriend}
\title{Friends black or white}
\source{
General Social Survey https://gss.norc.org
}
\description{
bwfriend
}
\details{
Question Are any of your good friends that you feel close to Black/White?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1136/vshow}{GSS Data Explorer page for \code{bwfriend}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 2385 \tab 204 \tab 241 \tab 2 \tab 2832 \cr
   Total \tab 2385 \tab 204 \tab 241 \tab 2 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numbwfrd}},
\code{\link{numfrend}}
}
\concept{Religion}
\keyword{variable}
