% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{byemail}
\alias{byemail}
\title{How many people send email}
\source{
General Social Survey https://gss.norc.org
}
\description{
byemail
}
\details{
Question Of these____friends and relatives, about how many do you stay in contact with by:
E. Communicating through electronic mail?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1 - 2
\item \code{3} 3 - 5
\item \code{4} 6 - 10
\item \code{5} 11 - 15
\item \code{6} 16-25
\item \code{7} 26-50
\item \code{8} 50+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2548/vshow}{GSS Data Explorer page for \code{byemail}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 - 2 \tab 3 - 5 \tab 6 - 10 \tab 11 - 15 \tab 16-25 \tab 26-50 \tab 50+ \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1884 \tab 533 \tab 76 \tab 84 \tab 84 \tab 42 \tab 44 \tab 22 \tab 18 \tab 1 \tab 29 \tab 2817 \cr
   2002 \tab 74 \tab 1278 \tab 233 \tab 348 \tab 310 \tab 169 \tab 160 \tab 126 \tab 53 \tab 1 \tab 13 \tab 2765 \cr
   2004 \tab 1873 \tab 363 \tab 117 \tab 142 \tab 139 \tab 49 \tab 52 \tab 50 \tab 15 \tab - \tab 12 \tab 2812 \cr
   Total \tab 3831 \tab 2174 \tab 426 \tab 574 \tab 533 \tab 260 \tab 256 \tab 198 \tab 86 \tab 2 \tab 54 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}},
\code{\link{polinfgn}}
}
\concept{Information Society}
\keyword{variable}
