% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{byemprob}
\alias{byemprob}
\title{How many friends discuss problems via email}
\source{
General Social Survey https://gss.norc.org
}
\description{
byemprob
}
\details{
Question A. Of these (NUMPROBS) friends and relatives that you feel really close to, how many do you stay in contact with  through electronic mail or instant messaging?
}
\section{Values}{

\itemize{
\item \verb{[96]} 96 or higher
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2607/vshow}{GSS Data Explorer page for \code{byemprob}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 17 \tab 18 \tab 2 \tab 20 \tab 25 \tab 26 \tab 3 \tab 30 \tab 35 \tab 4 \tab 40 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab 96 or higher \tab don't know \tab no answer \tab Total \cr
   2002 \tab 179 \tab 1318 \tab 164 \tab 130 \tab 6 \tab 17 \tab 3 \tab 2 \tab 55 \tab 2 \tab 3 \tab 234 \tab 32 \tab 21 \tab 1 \tab 159 \tab 9 \tab 2 \tab 93 \tab 2 \tab 187 \tab 8 \tab 55 \tab 2 \tab 23 \tab 37 \tab 4 \tab 2 \tab 1 \tab 14 \tab 2765 \cr
   Total \tab 179 \tab 1318 \tab 164 \tab 130 \tab 6 \tab 17 \tab 3 \tab 2 \tab 55 \tab 2 \tab 3 \tab 234 \tab 32 \tab 21 \tab 1 \tab 159 \tab 9 \tab 2 \tab 93 \tab 2 \tab 187 \tab 8 \tab 55 \tab 2 \tab 23 \tab 37 \tab 4 \tab 2 \tab 1 \tab 14 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{intaccss}},
\code{\link{jobinfo}},
\code{\link{jobpaper}},
\code{\link{numprobs}}
}
\concept{Information Society}
\keyword{variable}
