% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{byphone}
\alias{byphone}
\title{How many people talk to on phone}
\source{
General Social Survey https://gss.norc.org
}
\description{
byphone
}
\details{
Question Of these____friends and relatives, about how many do you stay in contact with by:
B. Talking with them on the telephone?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1 - 2
\item \code{3} 3 - 5
\item \code{4} 6 - 10
\item \code{5} 11 - 15
\item \code{6} 16-25
\item \code{7} 26-50
\item \code{8} 50+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2545/vshow}{GSS Data Explorer page for \code{byphone}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 - 2 \tab 3 - 5 \tab 6 - 10 \tab 11 - 15 \tab 16-25 \tab 26-50 \tab 50+ \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1884 \tab 38 \tab 95 \tab 236 \tab 227 \tab 102 \tab 116 \tab 63 \tab 33 \tab 1 \tab 22 \tab 2817 \cr
   2002 \tab 74 \tab 69 \tab 262 \tab 608 \tab 634 \tab 336 \tab 380 \tab 275 \tab 113 \tab - \tab 14 \tab 2765 \cr
   2004 \tab 1873 \tab 21 \tab 137 \tab 188 \tab 227 \tab 109 \tab 133 \tab 83 \tab 29 \tab - \tab 12 \tab 2812 \cr
   Total \tab 3831 \tab 128 \tab 494 \tab 1032 \tab 1088 \tab 547 \tab 629 \tab 421 \tab 175 \tab 1 \tab 48 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}},
\code{\link{polinfgn}}
}
\concept{Information Society}
\keyword{variable}
