% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{camwrk}
\alias{camwrk}
\title{Govt camera surveillance at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
camwrk
}
\details{
Question Protecting secrets is a continuing concern of the government. I am going to ask you about various security measures that the government might apply to individuals with a SECRET or TOP SECRET clearance. Do you (strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree) that people with SECRET or TOP SECRET clearance should be subject to:
H. Video camera surveillance in the workplace
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5256/vshow}{GSS Data Explorer page for \code{camwrk}}.
}

\seealso{
Other National Security: 
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{nuclrwar}},
\code{\link{srchwrk}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{usspy}},
\code{\link{usterror}}

Other United States: 
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{nuclrwar}},
\code{\link{srchwrk}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{usspy}},
\code{\link{usterror}}
}
\concept{National Security}
\concept{United States}
\keyword{variable}
