% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{caninf14}
\alias{caninf14}
\title{First site mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
caninf14
}
\details{
Question Imagine you wanted to use the Web to find out information about a candidate for political office. What would you do?
N. First site mentioned.
}
\section{Values}{

\itemize{
\item \code{0} no specific search engine mentioned
\item \code{1} yahoo
\item \code{2} aol
\item \code{3} cnn
\item \code{4} msn
\item \code{5} excite
\item \code{6} google
\item \code{7} webferret
\item \code{8} lycos
\item \code{9} altavista
\item \code{10} usatoday
\item \code{11} copernic
\item \code{12} askjeeves
\item \code{13} peoplefind.com
\item \code{14} abc
\item \code{15} msnbc
\item \code{16} infoseek go
\item \code{17} dogpile
\item \code{18} snap
\item \code{19} metacrawler
\item \code{20} webcrawler
\item \code{21} netscape
\item \code{22} iwon
\item \code{23} politics.com
\item \code{24} homp page search engine
\item \code{25} nbc
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2997/vshow}{GSS Data Explorer page for \code{caninf14}}.
}

\seealso{
Other Information Seeking Online: 
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf11}},
\code{\link{caninf12}},
\code{\link{caninf13}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf6}},
\code{\link{caninf7}},
\code{\link{caninf8}},
\code{\link{caninf9}}
}
\concept{Information Seeking Online}
\keyword{variable}
