% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_334.R
\name{caninf6}
\alias{caninf6}
\title{Would use chat group for information on candida}
\source{
General Social Survey https://gss.norc.org
}
\description{
caninf6
}
\details{
Question Imagine you wanted to use the Web to find out information about a candidate for political office. What would you do?
F. Would use chat group, news group, discussion forum or other online interactive service for information on candidate.
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{2} yes, at least one url is not a .com
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2989/vshow}{GSS Data Explorer page for \code{caninf6}}.
}

\seealso{
Other Information Seeking Online: 
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf7}},
\code{\link{caninf8}},
\code{\link{caninf9}}

Other Political: 
\code{\link{caninf1}},
\code{\link{caninf10}},
\code{\link{caninf2}},
\code{\link{caninf3}},
\code{\link{caninf4}},
\code{\link{caninf5}},
\code{\link{caninf7}},
\code{\link{caninf8}},
\code{\link{caninf9}}
}
\concept{Information Seeking Online}
\concept{Political}
\keyword{variable}
