% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{cantrust}
\alias{cantrust}
\title{People can be trusted or cant be too careful}
\source{
General Social Survey https://gss.norc.org
}
\description{
cantrust
}
\details{
Question Generally speaking, would you say that people can be trusted or that you can't be too careful in dealing with people?
}
\section{Values}{

\itemize{
\item \code{1} people can almost always be trusted
\item \code{2} people can usually be trusted
\item \code{3} you usually can't be too careful in dealing with people
\item \code{4} you almost always can't be too careful in dealing with people
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4686/vshow}{GSS Data Explorer page for \code{cantrust}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{socright}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
