% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{capfirm}
\alias{capfirm}
\title{How firm is r's opinion of death penalty}
\source{
General Social Survey https://gss.norc.org
}
\description{
capfirm
}
\details{
Question How firm are you about your opinion on the death penalty--would you say you are very likely to change your opinion, somewhat likely to change, somewhat unlikely to change, or very unlikely to change?
}
\section{Values}{

\itemize{
\item \code{1} change very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} change very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/271/vshow}{GSS Data Explorer page for \code{capfirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab change very likely \tab somewhat likely \tab somewhat unlikely \tab change very unlikely \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 13 \tab 95 \tab 141 \tab 216 \tab 15 \tab 4 \tab 1473 \cr
   Total \tab 989 \tab 13 \tab 95 \tab 141 \tab 216 \tab 15 \tab 4 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
