% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{capimp}
\alias{capimp}
\title{Importance of death penalty issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
capimp
}
\details{
Question How important is the death penalty issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most imp
\item \verb{[2]} important
\item \verb{[3]} not very imp
\item \verb{[4]} not important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/269/vshow}{GSS Data Explorer page for \code{capimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not important \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 67 \tab 261 \tab 120 \tab 26 \tab 6 \tab 4 \tab 1473 \cr
   Total \tab 989 \tab 67 \tab 261 \tab 120 \tab 26 \tab 6 \tab 4 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}
}
\concept{Core}
\keyword{variable}
