% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{cappun2}
\alias{cappun2}
\title{Does r favor capital punishment}
\source{
General Social Survey https://gss.norc.org
}
\description{
cappun2
}
\details{
Question Are you in favor of the death penalty for persons convicted of murder?
}
\section{Values}{

\itemize{
\item \code{1} yes,favor
\item \code{2} no,dont favor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
   1973 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/267/vshow}{GSS Data Explorer page for \code{cappun2}}.
}

\seealso{
Other Core: 
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}
}
\concept{Core}
\keyword{variable}
