% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{cardo}
\alias{cardo}
\title{What would you do in this situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
cardo
}
\details{
Question Suppose you were riding in a car driven by a close friend. You know he is going too fast. He hits a pedestrian. He asks you to tell the police that he was obeying the speed limit.
B. What would you do in this situation?
}
\section{Values}{

\itemize{
\item \verb{[1]} def tll plce was
\item \verb{[2]} prb tll plce was
\item \verb{[3]} prb tll plce not
\item \verb{[4]} def tll plce not
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4697/vshow}{GSS Data Explorer page for \code{cardo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab def tll plce was \tab prb tll plce was \tab prb tll plce not \tab def tll plce not \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 443 \tab 415 \tab 103 \tab 17 \tab 245 \tab 61 \tab 2832 \cr
   Total \tab 1548 \tab 443 \tab 415 \tab 103 \tab 17 \tab 245 \tab 61 \tab 2832 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{carright}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{ISSP Religion}
\keyword{variable}
