% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{carecost}
\alias{carecost}
\title{Who should pay for costs of care for under school age child}
\source{
General Social Survey https://gss.norc.org
}
\description{
carecost
}
\details{
Question Who do you think should primarily cover the costs of childcare for children under school age?
}
\section{Values}{

\itemize{
\item \verb{[1]} the family
\item \verb{[2]} the government/public funds
\item \verb{[3]} the employers
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4470/vshow}{GSS Data Explorer page for \code{carecost}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab the family \tab the government/public funds \tab the employers \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 846 \tab 167 \tab 73 \tab 208 \tab 8 \tab - \tab 1974 \cr
   2022 \tab 2384 \tab 520 \tab 281 \tab 145 \tab 200 \tab 5 \tab 9 \tab 3544 \cr
   Total \tab 3056 \tab 1366 \tab 448 \tab 218 \tab 408 \tab 13 \tab 9 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
