% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{carecost}
\alias{carecost}
\title{Who should pay for costs of care for under school age child}
\source{
General Social Survey https://gss.norc.org
}
\description{
carecost
}
\details{
Question Who do you think should primarily cover the costs of childcare for children under school age?
}
\section{Values}{

\itemize{
\item \code{1} the family
\item \code{2} the government/public funds
\item \code{3} the employers
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4470/vshow}{GSS Data Explorer page for \code{carecost}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}

Other Spending: 
\code{\link{careprov}}
}
\concept{ISSP Family & Gender Roles}
\concept{Spending}
\keyword{variable}
