% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{caregrp}
\alias{caregrp}
\title{How committed to group is r}
\source{
General Social Survey https://gss.norc.org
}
\description{
caregrp
}
\details{
Question Personally, how committed do you feel to this organization?  Not committed at all, slightly committed, moderately committed, strongly committed, or very strongly committed.
}
\section{Values}{

\itemize{
\item \verb{[1]} not committed
\item \verb{[2]} slightly comm
\item \verb{[3]} moderately comm
\item \verb{[4]} strongly comm
\item \verb{[5]} very strongly comm
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1014/vshow}{GSS Data Explorer page for \code{caregrp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not committed \tab slightly comm \tab moderately comm \tab strongly comm \tab very strongly comm \tab don't know \tab no answer \tab Total \cr
   1987 \tab 606 \tab 50 \tab 170 \tab 275 \tab 351 \tab 331 \tab 1 \tab 35 \tab 1819 \cr
   Total \tab 606 \tab 50 \tab 170 \tab 275 \tab 351 \tab 331 \tab 1 \tab 35 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{churhgrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{mostact}},
\code{\link{servegrp}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
