% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{careprov}
\alias{careprov}
\title{Who should provide childcare for under school age child}
\source{
General Social Survey https://gss.norc.org
}
\description{
careprov
}
\details{
Question People have different views on childcare for children under school age. Who do you think should primarily provide childcare?
}
\section{Values}{

\itemize{
\item \verb{[1]} family members
\item \verb{[2]} government agencies
\item \verb{[3]} non-profit organizations (e.g., charitable organizations, churches/religious organizations)
\item \verb{[4]} private childcare providers (e.g., private day care center, nanny, babysitter)
\item \verb{[5]} employers
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4469/vshow}{GSS Data Explorer page for \code{careprov}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab family members \tab government agencies \tab non-profit organizations (e.g., charitable organizations, churches/religious organizations) \tab private childcare providers (e.g., private day care center, nanny, babysitter) \tab employers \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 623 \tab 86 \tab 39 \tab 279 \tab 76 \tab 191 \tab 8 \tab - \tab 1974 \cr
   2022 \tab 2384 \tab 450 \tab 109 \tab 29 \tab 214 \tab 108 \tab 234 \tab 6 \tab 10 \tab 3544 \cr
   Total \tab 3056 \tab 1073 \tab 195 \tab 68 \tab 493 \tab 184 \tab 425 \tab 14 \tab 10 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
