% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{careself}
\alias{careself}
\title{Those in need have to take care of themselves}
\source{
General Social Survey https://gss.norc.org
}
\description{
careself
}
\details{
Question Please tell me whether you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree with the following statements:
B. Those in need have to learn to take care of themselves and not depend on others
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2012 \tab A/C/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2891/vshow}{GSS Data Explorer page for \code{careself}}.
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selffrst}},
\code{\link{selfless}}
}
\concept{Altruism}
\keyword{variable}
