% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{carsfam}
\alias{carsfam}
\title{Car pollution danger to my family}
\source{
General Social Survey https://gss.norc.org
}
\description{
carsfam
}
\details{
Question Please check one box for each of the questions below to show what you think.
B. And do you think that air pollution caused by cars is…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4754/vshow}{GSS Data Explorer page for \code{carsfam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab Total \cr
   1993 \tab 49 \tab 189 \tab 372 \tab 674 \tab 178 \tab 22 \tab 36 \tab 86 \tab 1606 \cr
   1994 \tab 1606 \tab 161 \tab 314 \tab 613 \tab 163 \tab 23 \tab 42 \tab 70 \tab 2992 \cr
   2000 \tab 1541 \tab 175 \tab 329 \tab 560 \tab 102 \tab 8 \tab 46 \tab 56 \tab 2817 \cr
   Total \tab 3196 \tab 525 \tab 1015 \tab 1847 \tab 443 \tab 53 \tab 124 \tab 212 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
