% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{carsten}
\alias{carsten}
\title{Car pollution will increase in next 10 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
carsten
}
\details{
Question Please check one box for each of the questions below to show what you think.
C. Within the next ten years, how likely do you think it is that there will be a large increase in ill-health In America's cities as a result of air pollution caused by cars?
}
\section{Values}{

\itemize{
\item \code{1} certain to happen
\item \code{2} very likely to happen
\item \code{3} fairly likely to happen
\item \code{4} not very likely to happen
\item \code{5} not at all likely
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4755/vshow}{GSS Data Explorer page for \code{carsten}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
