% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_313.R
\name{cesd1}
\alias{cesd1}
\title{How much time felt depressed in past wk}
\source{
General Social Survey https://gss.norc.org
}
\description{
cesd1
}
\details{
Question (HAND CARD N1)
I will now read out a list of the ways you might have felt or behaved during the past week. Using this card, please tell me how much of the time during the past week...                       
...you felt depressed?
}
\section{Values}{

\itemize{
\item \verb{[1]} none or almost none of the time
\item \verb{[2]} some of the time
\item \verb{[3]} most of the time
\item \verb{[4]} all or almost all of the time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6297/vshow}{GSS Data Explorer page for \code{cesd1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab none or almost none of the time \tab some of the time \tab most of the time \tab all or almost all of the time \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1888 \tab 646 \tab 264 \tab 32 \tab 21 \tab 3 \tab 13 \tab 2867 \cr
   Total \tab 1888 \tab 646 \tab 264 \tab 32 \tab 21 \tab 3 \tab 13 \tab 2867 \cr
}
}

\seealso{
Other ESS Health: 
\code{\link{cesd2}},
\code{\link{cesd3}},
\code{\link{cesd4}},
\code{\link{cesd5}}
}
\concept{ESS Health}
\keyword{variable}
