% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{changed2}
\alias{changed2}
\title{Life changes from religious change2}
\source{
General Social Survey https://gss.norc.org
}
\description{
changed2
}
\details{
Question The consequences of the spiritual or religious change.
B. Second mention
}
\section{Values}{

\itemize{
\item \code{96} no mention
\item \code{97} nothing special happened
\item \code{99} no asnwer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3075/vshow}{GSS Data Explorer page for \code{changed2}}.
}

\seealso{
Other Spirituality and Religion: 
\code{\link{alloflfe}},
\code{\link{changed1}},
\code{\link{changed3}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}}

Other Religion: 
\code{\link{alloflfe}},
\code{\link{changed1}},
\code{\link{changed3}},
\code{\link{relalt1}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}}
}
\concept{Religion}
\concept{Spirituality and Religion}
\keyword{variable}
