% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{chatmin}
\alias{chatmin}
\title{Chatroom minutes per week}
\source{
General Social Survey https://gss.norc.org
}
\description{
chatmin
}
\details{
Question Earlier you mentioned that you spend (number in answer to WWWHR, WWWMIN) (minutes/hours) using the World Wide Web in a typical week. About how many of those do you spend using chat rooms, news groups, bulletin boards, discussion Forums, and other forms of on-line interaction with other Net users–not just browsing on your own?
B. Minutes per week?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2387/vshow}{GSS Data Explorer page for \code{chatmin}}.
}

\seealso{
Other Information Society: 
\code{\link{chathr}},
\code{\link{drctlink}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{srcheng}},
\code{\link{typeurl}},
\code{\link{usewww}},
\code{\link{wwwhr}},
\code{\link{wwwmin}}

Other Computers And Internet: 
\code{\link{chathr}},
\code{\link{drctlink}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{srcheng}},
\code{\link{typeurl}},
\code{\link{usewww}},
\code{\link{wwwhr}},
\code{\link{wwwmin}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
