% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chembal}
\alias{chembal}
\title{Situation by a chemical imbalance in the brain}
\source{
General Social Survey https://gss.norc.org
}
\description{
chembal
}
\details{
Question B. In your opinion, how likely is it that (John/Mary)''s situation MIGHT be caused by a chemical imbalance in the brain?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2922/vshow}{GSS Data Explorer page for \code{chembal}}.
}

\seealso{
Other Mental Health II: 
\code{\link{badchar}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldprob}},
\code{\link{chldvig}},
\code{\link{stressfl}}

Other Children: 
\code{\link{badchar}},
\code{\link{chldhlp}},
\code{\link{chldprob}},
\code{\link{chldvig}},
\code{\link{stressfl}}
}
\concept{Children}
\concept{Mental Health II}
\keyword{variable}
