% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{chemfam}
\alias{chemfam}
\title{Pesticides danger to my family}
\source{
General Social Survey https://gss.norc.org
}
\description{
chemfam
}
\details{
Question B. And do you think that pesticides and chemicals used in farming are…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4761/vshow}{GSS Data Explorer page for \code{chemfam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab Total \cr
   1993 \tab 49 \tab 211 \tab 328 \tab 736 \tab 194 \tab 18 \tab 40 \tab 30 \tab 1606 \cr
   1994 \tab 1606 \tab 186 \tab 265 \tab 640 \tab 188 \tab 21 \tab 39 \tab 47 \tab 2992 \cr
   Total \tab 1655 \tab 397 \tab 593 \tab 1376 \tab 382 \tab 39 \tab 79 \tab 77 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\keyword{variable}
