% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{chemgen}
\alias{chemgen}
\title{Pesticides danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
chemgen
}
\details{
Question A. In general, do you think that pesticides and chemicals used in farming are...
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4760/vshow}{GSS Data Explorer page for \code{chemgen}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 225 \tab 358 \tab 737 \tab 161 \tab 8 \tab 40 \tab 28 \tab 1606 \cr
   1994 \tab 1606 \tab 177 \tab 280 \tab 677 \tab 162 \tab 13 \tab 33 \tab 44 \tab 2992 \cr
   2000 \tab 1541 \tab 216 \tab 366 \tab 504 \tab 78 \tab 15 \tab 46 \tab 51 \tab 2817 \cr
   2010 \tab 614 \tab 277 \tab 464 \tab 525 \tab 103 \tab 7 \tab 47 \tab 7 \tab 2044 \cr
   2016 \tab 1956 \tab 263 \tab 288 \tab 283 \tab 49 \tab 12 \tab 14 \tab 2 \tab 2867 \cr
   2018 \tab 1563 \tab 242 \tab 248 \tab 244 \tab 37 \tab 5 \tab 6 \tab 3 \tab 2348 \cr
   Total \tab 7329 \tab 1400 \tab 2004 \tab 2970 \tab 590 \tab 60 \tab 186 \tab 135 \tab 14674 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{indusgen}},
\code{\link{tempgen}},
\code{\link{watergen}}

Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
