% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{chemgen1}
\alias{chemgen1}
\title{Pesticides danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
chemgen1
}
\details{
Question And do you think that pesticides and chemicals used in farming are…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7491/vshow}{GSS Data Explorer page for \code{chemgen1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 419 \tab 578 \tab 667 \tab 97 \tab 12 \tab 60 \tab 185 \tab 11 \tab 4032 \cr
   Total \tab 2003 \tab 419 \tab 578 \tab 667 \tab 97 \tab 12 \tab 60 \tab 185 \tab 11 \tab 4032 \cr
}
}

\keyword{variable}
