% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_668.R
\name{childs_exp}
\alias{childs_exp}
\title{Total children under 18 in hh (experimental verison)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childs_exp
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDS_EXP has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDS_EXP. The original CHILDS_EXP containing data in 2022 has been renamed CHILDS_EXP_22.
}
\section{Values}{

\itemize{
\item \code{8} 8 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7999/vshow}{GSS Data Explorer page for \code{childs_exp}}.

Counts by year:\tabular{llllllllllll}{
   year \tab NA \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab Total \cr
   2022 \tab 1802 \tab 1214 \tab 215 \tab 177 \tab 85 \tab 30 \tab 13 \tab 4 \tab 3 \tab 1 \tab 3544 \cr
   Total \tab 1802 \tab 1214 \tab 215 \tab 177 \tab 85 \tab 30 \tab 13 \tab 4 \tab 3 \tab 1 \tab 3544 \cr
}
}

\seealso{
Other Core: 
\code{\link{childsinhh}}
}
\concept{Core}
\keyword{variable}
