% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{childs_exp}
\alias{childs_exp}
\title{Total children under 18 in hh (experimental verison)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childs_exp
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDS_EXP has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDS_EXP. The original CHILDS_EXP containing data in 2022 has been renamed CHILDS_EXP_22.
}
\section{Values}{

\itemize{
\item \code{8} 8 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7999/vshow}{GSS Data Explorer page for \code{childs_exp}}.
}

\concept{Core}
\keyword{variable}
