% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_668.R
\name{childsinhh}
\alias{childsinhh}
\title{Total children in hh (household composition version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childsinhh
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDSINHH has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDSINHH. The original CHILDSINHH containing data from 2021 to 2022 has been renamed CHILDSINHH_2122.
}
\section{Values}{

\itemize{
\item \code{8} 8 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8006/vshow}{GSS Data Explorer page for \code{childsinhh}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 or more \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1764 \tab 1351 \tab 155 \tab 166 \tab 70 \tab 15 \tab 6 \tab 2 \tab 1 \tab 2 \tab 8 \tab 4 \tab 3544 \cr
   Total \tab 5796 \tab 1351 \tab 155 \tab 166 \tab 70 \tab 15 \tab 6 \tab 2 \tab 1 \tab 2 \tab 8 \tab 4 \tab 7576 \cr
}
}

\seealso{
Other Core: 
\code{\link{childs_exp}}
}
\concept{Core}
\keyword{variable}
