% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{childsinhh}
\alias{childsinhh}
\title{Total children in hh (household composition version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
childsinhh
}
\details{
Question Calculated variable: Total count of children under 18 in household
As of 2024, CHILDSINHH has been calculated to only account for 0 to 6+ children in the household. This has been retroactively applied to all previous years of CHILDSINHH. The original CHILDSINHH containing data from 2021 to 2022 has been renamed CHILDSINHH_2122.
}
\section{Values}{

\itemize{
\item \code{8} 8 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8006/vshow}{GSS Data Explorer page for \code{childsinhh}}.
}

\concept{Core}
\keyword{variable}
