% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{chkfinan}
\alias{chkfinan}
\title{Government should verify financial assets}
\source{
General Social Survey https://gss.norc.org
}
\description{
chkfinan
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should contact other people and verify information provided by the individual concerning his or her: (READ A-C)
A. Financial assets and liabilities
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5223/vshow}{GSS Data Explorer page for \code{chkfinan}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 693 \tab 420 \tab 174 \tab 123 \tab 48 \tab 2 \tab 2904 \cr
   1998 \tab 1445 \tab 538 \tab 454 \tab 214 \tab 121 \tab 49 \tab 11 \tab 2832 \cr
   Total \tab 2889 \tab 1231 \tab 874 \tab 388 \tab 244 \tab 97 \tab 13 \tab 5736 \cr
}
}

\seealso{
Other National Security: 
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{punarms}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}},
\code{\link{taketrck}}
}
\concept{National Security}
\keyword{variable}
