% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{chkspfin}
\alias{chkspfin}
\title{Government should verify spouse's financial assets}
\source{
General Social Survey https://gss.norc.org
}
\description{
chkspfin
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should contact other people and verify information provided by the individual concerning his or her: (READ A-C)
B. Spouse's financial assets and liabilities
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5224/vshow}{GSS Data Explorer page for \code{chkspfin}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 542 \tab 436 \tab 259 \tab 170 \tab 50 \tab 3 \tab 2904 \cr
   1998 \tab 1445 \tab 434 \tab 433 \tab 287 \tab 166 \tab 55 \tab 12 \tab 2832 \cr
   Total \tab 2889 \tab 976 \tab 869 \tab 546 \tab 336 \tab 105 \tab 15 \tab 5736 \cr
}
}

\seealso{
Other National Security: 
\code{\link{chkfinan}},
\code{\link{chktaxes}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{punarms}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}},
\code{\link{taketrck}}
}
\concept{National Security}
\keyword{variable}
