% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{chktaxes}
\alias{chktaxes}
\title{Government should verify tax records}
\source{
General Social Survey https://gss.norc.org
}
\description{
chktaxes
}
\details{
Question Before giving an individual a SECRET or TOP SECRET clearance, the government should contact other people and verify information provided by the individual concerning his or her: (READ A-C)
C. Tax records
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should
\item \verb{[2]} probably should
\item \verb{[3]} probably should not
\item \verb{[4]} definitely should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5225/vshow}{GSS Data Explorer page for \code{chktaxes}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 672 \tab 448 \tab 171 \tab 117 \tab 47 \tab 5 \tab 2904 \cr
   1998 \tab 1445 \tab 528 \tab 445 \tab 221 \tab 129 \tab 53 \tab 11 \tab 2832 \cr
   Total \tab 2889 \tab 1200 \tab 893 \tab 392 \tab 246 \tab 100 \tab 16 \tab 5736 \cr
}
}

\seealso{
Other National Security: 
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{punarms}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}},
\code{\link{taketrck}}
}
\concept{National Security}
\keyword{variable}
