% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_685.R
\name{chldadlt_wht}
\alias{chldadlt_wht}
\title{Percent of adults you knew (friends, care takers, neighbors) during childhood who were white}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldadlt_wht
}
\details{
Question Not including your family, what percentage of the adults you knew (friends, care takers, neighbors) belonged to each of the following categories? - White
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8075/vshow}{GSS Data Explorer page for \code{chldadlt_wht}}.
}

\seealso{
Other Mental Health: 
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_wht}},
\code{\link{chldfam_oth}}
}
\concept{Mental Health}
\keyword{variable}
