% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_558.R
\name{chldeduc}
\alias{chldeduc}
\title{Education improving?}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldeduc
}
\details{
Question Compared with 10 years ago, would you say that American children today are much better off, better off, about the same, worse off, or much worse off regarding:
A. The quality of their education.
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} better
\item \code{3} about the same
\item \code{4} worse
\item \code{5} much worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5108/vshow}{GSS Data Explorer page for \code{chldeduc}}.
}

\seealso{
Other Children: 
\code{\link{chldlove}},
\code{\link{chldmorl}},
\code{\link{chldsafe}},
\code{\link{chldsup}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolepa}},
\code{\link{roletchr}}
}
\concept{Children}
\keyword{variable}
