% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{chldfam_wht}
\alias{chldfam_wht}
\title{Percent of family during childhood who were white}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldfam_wht
}
\details{
Question What percentage of your family belong to each of the following categories? - White
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8085/vshow}{GSS Data Explorer page for \code{chldfam_wht}}.
}

\seealso{
Other Mental Health: 
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}}
}
\concept{Mental Health}
\keyword{variable}
