% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chldhlp}
\alias{chldhlp}
\title{What should be done to help a named youth}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldhlp
}
\details{
Question What (if anything) should be done to help (John/Mary)?
PROBE: Anything else?
}
\section{Values}{

\itemize{
\item \code{10} nothing
\item \code{20} medical general
\item \code{21} doctor
\item \code{22} specialist
\item \code{30} mental health general
\item \code{31} psychiatrist
\item \code{32} social worker
\item \code{33} counselor therapist
\item \code{34} psychologist
\item \code{40} family friends general
\item \code{41} negative blaming parenting
\item \code{42} positive parenting
\item \code{50} assistance from school general
\item \code{51} involve teacher
\item \code{52} involve school counselor
\item \code{60} religion
\item \code{80} social extracurricular activity
\item \code{90} other help miscellaneous
\item \code{97} dk ref
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2920/vshow}{GSS Data Explorer page for \code{chldhlp}}.
}

\seealso{
Other Mental Health II: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldprb}},
\code{\link{chldprob}},
\code{\link{chldvig}},
\code{\link{stressfl}}
}
\concept{Mental Health II}
\keyword{variable}
