% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{chldhome}
\alias{chldhome}
\title{Housing improving?}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldhome
}
\details{
Question Compared with 10 years ago, would you say that American children today are much better off, better off, about the same, worse off, or much worse off regarding:
F. Their housing.
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} better
\item \code{3} about the same
\item \code{4} worse
\item \code{5} much worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5113/vshow}{GSS Data Explorer page for \code{chldhome}}.
}

\seealso{
Other Children: 
\code{\link{chldhlth}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{sppregnt}}

Other Children: 
\code{\link{chldhlth}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{sppregnt}}
}
\concept{Children}
\keyword{variable}
