% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chldprb}
\alias{chldprb}
\title{What r think may be wrong with a named youth}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldprb
}
\details{
Question First, I'd like to know what you think may be wrong (if anything) with (John/Mary).
PROBE: Anything else?
}
\section{Values}{

\itemize{
\item \verb{[1]} adhd or add
\item \verb{[2]} depression
\item \verb{[3]} normal
\item \verb{[4]} asthma
\item \verb{[5]} developmental puberty
\item \verb{[6]} illicit drugs
\item \verb{[7]} adhd symptoms
\item \verb{[8]} psychiatric personality problems
\item \verb{[9]} medical problems symptoms
\item \verb{[10]} environment social
\item \verb{[11]} other characteristic miscellaneous
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2918/vshow}{GSS Data Explorer page for \code{chldprb}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab adhd or add \tab depression \tab normal \tab asthma \tab developmental puberty \tab illicit drugs \tab adhd symptoms \tab psychiatric personality problems \tab medical problems symptoms \tab environment social \tab other characteristic miscellaneous \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 135 \tab 228 \tab 287 \tab 230 \tab 30 \tab 49 \tab 43 \tab 61 \tab 62 \tab 161 \tab 26 \tab 78 \tab 3 \tab 2765 \cr
   Total \tab 1372 \tab 135 \tab 228 \tab 287 \tab 230 \tab 30 \tab 49 \tab 43 \tab 61 \tab 62 \tab 161 \tab 26 \tab 78 \tab 3 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldprob}},
\code{\link{chldvig}},
\code{\link{stressfl}}
}
\concept{Mental Health II}
\keyword{variable}
