% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chldprb}
\alias{chldprb}
\title{What r think may be wrong with a named youth}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldprb
}
\details{
Question First, I'd like to know what you think may be wrong (if anything) with (John/Mary).
PROBE: Anything else?
}
\section{Values}{

\itemize{
\item \code{1} adhd or add
\item \code{2} depression
\item \code{3} normal
\item \code{4} asthma
\item \code{5} developmental puberty
\item \code{6} illicit drugs
\item \code{7} adhd symptoms
\item \code{8} psychiatric personality problems
\item \code{9} medical problems symptoms
\item \code{10} environment social
\item \code{11} other characteristic miscellaneous
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2918/vshow}{GSS Data Explorer page for \code{chldprb}}.
}

\seealso{
Other Mental Health II: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldprob}},
\code{\link{chldvig}},
\code{\link{stressfl}}
}
\concept{Mental Health II}
\keyword{variable}
