% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_689.R
\name{chldpref_next}
\alias{chldpref_next}
\title{R ideal number of children - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldpref_next
}
\details{
Question For you personally, what is the ideal number of children you would like to have or would have liked to have had?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[7]} 7 or more
\item \verb{[8]} as many as you want
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8257/vshow}{GSS Data Explorer page for \code{chldpref_next}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab not available in this year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 or more \tab as many as you want \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 127 \tab 85 \tab 432 \tab 203 \tab 100 \tab 31 \tab 15 \tab 12 \tab 58 \tab 2244 \tab 2 \tab 3309 \cr
   Total \tab 0 \tab 127 \tab 85 \tab 432 \tab 203 \tab 100 \tab 31 \tab 15 \tab 12 \tab 58 \tab 2244 \tab 2 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{abhelp1}},
\code{\link{abhelp2}},
\code{\link{abhelp3}},
\code{\link{abhelp4}},
\code{\link{abmoral}},
\code{\link{advfront}},
\code{\link{biokids}},
\code{\link{blkdsrv}},
\code{\link{blkmblty}},
\code{\link{blktry}},
\code{\link{brv16}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16dadnum}},
\code{\link{brv16filnum}},
\code{\link{brv16frndnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16grand}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16oth}},
\code{\link{brv16othnum}},
\code{\link{brv16par}},
\code{\link{brv16partnum}},
\code{\link{brv16sib}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{brv16sp}},
\code{\link{brv16spnum}},
\code{\link{brv5}},
\code{\link{brv5bronum}},
\code{\link{brv5child}},
\code{\link{brv5chinum}},
\code{\link{brv5cowknum}},
\code{\link{brv5cuznum}},
\code{\link{brv5dadnum}},
\code{\link{brv5daunum}},
\code{\link{brv5filnum}},
\code{\link{brv5frndnum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5grand}},
\code{\link{brv5milnum}},
\code{\link{brv5momnum}},
\code{\link{brv5oth}},
\code{\link{brv5othnum}},
\code{\link{brv5par}},
\code{\link{brv5partnum}},
\code{\link{brv5sib}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}},
\code{\link{brv5sonnum}},
\code{\link{brv5sp}},
\code{\link{brv5spnum}},
\code{\link{chldidel_next}},
\code{\link{confedvacy}},
\code{\link{conpharvacy}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{famgen_next}},
\code{\link{femself}},
\code{\link{firstkidsex}},
\code{\link{fluvax}},
\code{\link{fund_next}},
\code{\link{gestate}},
\code{\link{happy_next}},
\code{\link{hhtype1_next}},
\code{\link{knwsexor}},
\code{\link{malekids}},
\code{\link{mascself}},
\code{\link{nexteligible}},
\code{\link{nextgen}},
\code{\link{nonbinkids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{perla_self}},
\code{\link{petsseq}},
\code{\link{scientbe}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{toofast}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}

Other Children: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adoption}},
\code{\link{adoptkid}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{agekdbrn}},
\code{\link{agekdbrn_9422}},
\code{\link{aidkids}},
\code{\link{allergic}},
\code{\link{amicable}},
\code{\link{amicably}},
\code{\link{amicablz}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{cathsch}},
\code{\link{chembal}},
\code{\link{childs}},
\code{\link{chldben}},
\code{\link{chldcare}},
\code{\link{chldeduc}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldhrt}},
\code{\link{chldidel}},
\code{\link{chldidel_next}},
\code{\link{chldlove}},
\code{\link{chldmore}},
\code{\link{chldmorl}},
\code{\link{chldnum}},
\code{\link{chldprob}},
\code{\link{chldsafe}},
\code{\link{chldsch}},
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{chldskls}},
\code{\link{chldsoon}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{chldvig}},
\code{\link{chrissch}},
\code{\link{clean}},
\code{\link{cleany}},
\code{\link{cleanz}},
\code{\link{comknows}},
\code{\link{consider}},
\code{\link{considey}},
\code{\link{considez}},
\code{\link{control}},
\code{\link{controly}},
\code{\link{controlz}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{discipln}},
\code{\link{eldersup}},
\code{\link{evkid}},
\code{\link{failure}},
\code{\link{famper4}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{fourkids}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{helpoth}},
\code{\link{hlth8}},
\code{\link{homesch}},
\code{\link{honest}},
\code{\link{honesty}},
\code{\link{honestz}},
\code{\link{illphys}},
\code{\link{impkids}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{interest}},
\code{\link{interesy}},
\code{\link{interesz}},
\code{\link{judgment}},
\code{\link{judgmeny}},
\code{\link{judgmenz}},
\code{\link{kd1jwoth}},
\code{\link{kd1relig}},
\code{\link{kd2jwoth}},
\code{\link{kd2relig}},
\code{\link{kd3jwoth}},
\code{\link{kd3relig}},
\code{\link{kd4jwoth}},
\code{\link{kd4relig}},
\code{\link{kd5jwoth}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{kdalive1}},
\code{\link{kdalive2}},
\code{\link{kdalive3}},
\code{\link{kdalive4}},
\code{\link{kdalive5}},
\code{\link{kdalive6}},
\code{\link{kdalive7}},
\code{\link{kdalive8}},
\code{\link{kdalive9}},
\code{\link{kdeduc1}},
\code{\link{kdeduc2}},
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdpicked}},
\code{\link{kdpres80}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel8}},
\code{\link{kdrel9}},
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{kdsex9}},
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{kdwork1}},
\code{\link{kdwork2}},
\code{\link{kdwrkslf}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}},
\code{\link{kdyrbrn9}},
\code{\link{kid5up}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidhlth}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidlived}},
\code{\link{kidneedy}},
\code{\link{kidnofre}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidnum}},
\code{\link{kidout}},
\code{\link{kidpars}},
\code{\link{kidresp}},
\code{\link{kidskips}},
\code{\link{kidsocst}},
\code{\link{kidssol}},
\code{\link{kidtrble}},
\code{\link{kidxfilm}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{knowkid3}},
\code{\link{knowkid4}},
\code{\link{law7}},
\code{\link{manners}},
\code{\link{mannersy}},
\code{\link{mannersz}},
\code{\link{mapaid}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{mntlill}},
\code{\link{morextrm}},
\code{\link{mykids}},
\code{\link{nextdoor}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{obey}},
\code{\link{obeys}},
\code{\link{obeysy}},
\code{\link{obeysz}},
\code{\link{obeythnk}},
\code{\link{obrespct}},
\code{\link{onekid}},
\code{\link{oth3}},
\code{\link{othkids}},
\code{\link{othsch}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{popular}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{raised}},
\code{\link{relsch}},
\code{\link{responsi}},
\code{\link{responsy}},
\code{\link{responsz}},
\code{\link{role}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{roley}},
\code{\link{rolez}},
\code{\link{sellbaby}},
\code{\link{spanking}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spendeve}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid4}},
\code{\link{spwrkpar}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{stressfl}},
\code{\link{studious}},
\code{\link{studiouy}},
\code{\link{studiouz}},
\code{\link{success}},
\code{\link{successy}},
\code{\link{successz}},
\code{\link{sufadult}},
\code{\link{thnkself}},
\code{\link{threkids}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}},
\code{\link{trbllaw}},
\code{\link{twokids}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{workhard}},
\code{\link{zombies}}
}
\concept{Children}
\concept{GSS Next}
\keyword{variable}
