% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chldprob}
\alias{chldprob}
\title{How serious r consider his or her problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldprob
}
\details{
Question How serious would you consider (John/Mary)'s problem, if any, to be: very serious, somewhat serious, not very serious, or not at all serious?
}
\section{Values}{

\itemize{
\item \code{1} very serious
\item \code{2} somewhat serious
\item \code{3} not very serious
\item \code{4} not at all serious
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2919/vshow}{GSS Data Explorer page for \code{chldprob}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very serious \tab somewhat serious \tab not very serious \tab not at all serious \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 625 \tab 389 \tab 168 \tab 171 \tab 28 \tab 12 \tab 2765 \cr
   Total \tab 1372 \tab 625 \tab 389 \tab 168 \tab 171 \tab 28 \tab 12 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldvig}},
\code{\link{stressfl}}
}
\concept{Mental Health II}
\keyword{variable}
