% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_329.R
\name{chldsch}
\alias{chldsch}
\title{How willing r have child in same class of r's ch}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldsch
}
\details{
Question D. How willing would you be to have (John/Mary) in your child's class at school?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely willing
\item \verb{[2]} probably willing
\item \verb{[3]} probably unwilling
\item \verb{[4]} definitely unwilling
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2936/vshow}{GSS Data Explorer page for \code{chldsch}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 602 \tab 620 \tab 98 \tab 32 \tab 28 \tab 13 \tab 2765 \cr
   Total \tab 1372 \tab 602 \tab 620 \tab 98 \tab 32 \tab 28 \tab 13 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{chldfrnd}},
\code{\link{spendeve}}
}
\concept{Mental Health II}
\keyword{variable}
