% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{chldsex}
\alias{chldsex}
\title{Qualities for boys or girls}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldsex
}
\details{
Question A. When you rated the importance of various qualities for children, were you thinking mostly about boys, mostly about girls, or about both boys and girls equally?
}
\section{Values}{

\itemize{
\item \code{1} mostly boys
\item \code{2} mostly girls
\item \code{3} both equally
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1980 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/518/vshow}{GSS Data Explorer page for \code{chldsex}}.
}

\seealso{
Other Core: 
\code{\link{chldsex1}},
\code{\link{helpoth}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socrel}},
\code{\link{thnkself}},
\code{\link{workhard}},
\code{\link{youngen}}
}
\concept{Core}
\keyword{variable}
