% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{chldshop_oth}
\alias{chldshop_oth}
\title{Percent of people who shopped at the supermarket during childhood who were some other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldshop_oth
}
\details{
Question Think about the supermarket where your family shopped. What percentage of the people who shopped at the supermarket belong to each of the following categories? - Some other race
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8089/vshow}{GSS Data Explorer page for \code{chldshop_oth}}.
}

\seealso{
Other Mental Health: 
\code{\link{chldchld_oth}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_wht}}
}
\concept{Mental Health}
\keyword{variable}
