% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{chldvig}
\alias{chldvig}
\title{Vignette asked}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldvig
}
\details{
Question Child mental health vignette version number
}
\section{Values}{

\itemize{
\item \verb{[1]} 1:  adhd, john, white, 8
\item \verb{[2]} 2:  adhd, john, white, 14
\item \verb{[3]} 3:  adhd, mary, white, 8
\item \verb{[4]} 4:  adhd, mary, white, 14
\item \verb{[5]} 5:  adhd, john, black, 8
\item \verb{[6]} 6:  adhd, john, black, 14
\item \verb{[7]} 7:  adhd, mary, black, 8
\item \verb{[8]} 8:  adhd, mary, black, 14
\item \verb{[9]} 9:  depression, john, white, 8
\item \verb{[10]} 10: depression, john, white, 14
\item \verb{[11]} 11: depression, mary, white, 8
\item \verb{[12]} 12: depression, mary, white, 14
\item \verb{[13]} 13: depression, john, black, 8
\item \verb{[14]} 14: depression, john, black, 14
\item \verb{[15]} 15: depression, mary, black, 8
\item \verb{[16]} 16: depression, mary, black, 14
\item \verb{[17]} 17: normal, john, white, 8
\item \verb{[18]} 18: normal, john, white, 14
\item \verb{[19]} 19: normal, mary, white, 8
\item \verb{[20]} 20: normal, mary, white, 14
\item \verb{[21]} 21: normal, john, black, 8
\item \verb{[22]} 22: normal, john, black, 14
\item \verb{[23]} 23: normal, mary, black, 8
\item \verb{[24]} 24: normal, mary, black, 14
\item \verb{[25]} 25: asthma, john, white, 8
\item \verb{[26]} 26: asthma, john, white, 14
\item \verb{[27]} 27: asthma, mary, white, 8
\item \verb{[28]} 28: asthma, mary, white, 14
\item \verb{[29]} 29: asthma, john, black, 8
\item \verb{[30]} 30: asthma, john, black, 14
\item \verb{[31]} 31: asthma, mary, black, 8
\item \verb{[32]} 32: asthma, mary, black, 14
\item \verb{[33]} 33: classic depression, john, white, 8
\item \verb{[34]} 34: classic depression, john, white, 14
\item \verb{[35]} 35: classic depression, mary, white, 8
\item \verb{[36]} 36: classic depression, mary, white, 14
\item \verb{[37]} 37: classic depression, john, black, 8
\item \verb{[38]} 38: classic depression, john, black, 14
\item \verb{[39]} 39: classic depression, mary, black, 8
\item \verb{[40]} 40: classic depression, mary, black, 14
\item \verb{[41]} 41: impulsivity, john, white, 8
\item \verb{[42]} 42: impulsivity, john, white, 14
\item \verb{[43]} 43: impulsivity, mary, white, 8
\item \verb{[44]} 44: impulsivity, mary, white, 14
\item \verb{[45]} 45: impulsivity, john, black, 8
\item \verb{[46]} 46: impulsivity, john, black, 14
\item \verb{[47]} 47: impulsivity, mary, black, 8
\item \verb{[48]} 48: impulsivity, mary, black, 14
\item \verb{[49]} 49: perfectionism, john, white, 8
\item \verb{[50]} 50: perfectionism, john, white, 14
\item \verb{[51]} 51: perfectionism, mary, white, 8
\item \verb{[52]} 52: perfectionism, mary, white, 14
\item \verb{[53]} 53: perfectionism, john, black, 8
\item \verb{[54]} 54: perfectionism, john, black, 14
\item \verb{[55]} 55: perfectionism, mary, black, 8
\item \verb{[56]} 56: perfectionism, mary, black, 14
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2917/vshow}{GSS Data Explorer page for \code{chldvig}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1:  adhd, john, white, 8 \tab 2:  adhd, john, white, 14 \tab 3:  adhd, mary, white, 8 \tab 4:  adhd, mary, white, 14 \tab 5:  adhd, john, black, 8 \tab 6:  adhd, john, black, 14 \tab 7:  adhd, mary, black, 8 \tab 8:  adhd, mary, black, 14 \tab 9:  depression, john, white, 8 \tab 10: depression, john, white, 14 \tab 11: depression, mary, white, 8 \tab 12: depression, mary, white, 14 \tab 13: depression, john, black, 8 \tab 14: depression, john, black, 14 \tab 15: depression, mary, black, 8 \tab 16: depression, mary, black, 14 \tab 17: normal, john, white, 8 \tab 18: normal, john, white, 14 \tab 19: normal, mary, white, 8 \tab 20: normal, mary, white, 14 \tab 21: normal, john, black, 8 \tab 22: normal, john, black, 14 \tab 23: normal, mary, black, 8 \tab 24: normal, mary, black, 14 \tab 25: asthma, john, white, 8 \tab 26: asthma, john, white, 14 \tab 27: asthma, mary, white, 8 \tab 28: asthma, mary, white, 14 \tab 29: asthma, john, black, 8 \tab 30: asthma, john, black, 14 \tab 31: asthma, mary, black, 8 \tab 32: asthma, mary, black, 14 \tab not available in this year \tab 33: classic depression, john, white, 8 \tab 34: classic depression, john, white, 14 \tab 35: classic depression, mary, white, 8 \tab 36: classic depression, mary, white, 14 \tab 37: classic depression, john, black, 8 \tab 38: classic depression, john, black, 14 \tab 39: classic depression, mary, black, 8 \tab 40: classic depression, mary, black, 14 \tab 41: impulsivity, john, white, 8 \tab 42: impulsivity, john, white, 14 \tab 43: impulsivity, mary, white, 8 \tab 44: impulsivity, mary, white, 14 \tab 45: impulsivity, john, black, 8 \tab 46: impulsivity, john, black, 14 \tab 47: impulsivity, mary, black, 8 \tab 48: impulsivity, mary, black, 14 \tab 49: perfectionism, john, white, 8 \tab 50: perfectionism, john, white, 14 \tab 51: perfectionism, mary, white, 8 \tab 52: perfectionism, mary, white, 14 \tab 53: perfectionism, john, black, 8 \tab 54: perfectionism, john, black, 14 \tab 55: perfectionism, mary, black, 8 \tab 56: perfectionism, mary, black, 14 \tab Total \cr
   2002 \tab 1372 \tab 36 \tab 45 \tab 39 \tab 47 \tab 42 \tab 45 \tab 47 \tab 39 \tab 42 \tab 40 \tab 50 \tab 43 \tab 54 \tab 48 \tab 48 \tab 49 \tab 40 \tab 49 \tab 48 \tab 42 \tab 48 \tab 44 \tab 38 \tab 43 \tab 47 \tab 44 \tab 34 \tab 41 \tab 36 \tab 38 \tab 46 \tab 41 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2024 \tab 1639 \tab 31 \tab 33 \tab 29 \tab 29 \tab 26 \tab 39 \tab 37 \tab 37 \tab 39 \tab 24 \tab 40 \tab 38 \tab 38 \tab 40 \tab 38 \tab 28 \tab 29 \tab 29 \tab 42 \tab 33 \tab 42 \tab 35 \tab 27 \tab 42 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 31 \tab 29 \tab 41 \tab 34 \tab 35 \tab 31 \tab 37 \tab 40 \tab 32 \tab 34 \tab 33 \tab 43 \tab 25 \tab 31 \tab 44 \tab 26 \tab 44 \tab 34 \tab 37 \tab 32 \tab 42 \tab 42 \tab 33 \tab 35 \tab 3309 \cr
   Total \tab 3011 \tab 67 \tab 78 \tab 68 \tab 76 \tab 68 \tab 84 \tab 84 \tab 76 \tab 81 \tab 64 \tab 90 \tab 81 \tab 92 \tab 88 \tab 86 \tab 77 \tab 69 \tab 78 \tab 90 \tab 75 \tab 90 \tab 79 \tab 65 \tab 85 \tab 47 \tab 44 \tab 34 \tab 41 \tab 36 \tab 38 \tab 46 \tab 41 \tab 0 \tab 31 \tab 29 \tab 41 \tab 34 \tab 35 \tab 31 \tab 37 \tab 40 \tab 32 \tab 34 \tab 33 \tab 43 \tab 25 \tab 31 \tab 44 \tab 26 \tab 44 \tab 34 \tab 37 \tab 32 \tab 42 \tab 42 \tab 33 \tab 35 \tab 6074 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldprob}},
\code{\link{stressfl}}

Other Children: 
\code{\link{badchar}},
\code{\link{chembal}},
\code{\link{chldhlp}},
\code{\link{chldprob}},
\code{\link{stressfl}}
}
\concept{Children}
\concept{Mental Health}
\concept{Mental Health II}
\keyword{variable}
