% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_143.R
\name{chngeorg}
\alias{chngeorg}
\title{How likely i'll try to find job in another org.}
\source{
General Social Survey https://gss.norc.org
}
\description{
chngeorg
}
\details{
Question All in all, how likely is it that you will try hard to find a job with another organization within the next 12 months?  Would you say you are not at all likely, somewhat likely, or very likely?
}
\section{Values}{

\itemize{
\item \code{1} not at all likely
\item \code{2} somewaht likely
\item \code{3} very likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1348/vshow}{GSS Data Explorer page for \code{chngeorg}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab not at all likely \tab somewaht likely \tab very likely \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 570 \tab 176 \tab 143 \tab 12 \tab 11 \tab 1517 \cr
   Total \tab 605 \tab 570 \tab 176 \tab 143 \tab 12 \tab 11 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{helporg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{samevals}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{wrkmuch}}
}
\concept{Work Organizations}
\keyword{variable}
