% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{chngewrk}
\alias{chngewrk}
\title{I would change my work for something different}
\source{
General Social Survey https://gss.norc.org
}
\description{
chngewrk
}
\details{
Question (HAND CARD D15)
Now think of the type of work you do, irrespective of the firm or organization you work for.
To what extent do you agree or disagree with each of the following statements?
Given the chance, I would change my present type of work for something different.
Would you say...
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6334/vshow}{GSS Data Explorer page for \code{chngewrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 121 \tab 253 \tab 175 \tab 270 \tab 113 \tab 2 \tab 6 \tab 2867 \cr
   Total \tab 1927 \tab 121 \tab 253 \tab 175 \tab 270 \tab 113 \tab 2 \tab 6 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
