% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{chngewrk}
\alias{chngewrk}
\title{I would change my work for something different}
\source{
General Social Survey https://gss.norc.org
}
\description{
chngewrk
}
\details{
Question (HAND CARD D15)
Now think of the type of work you do, irrespective of the firm or organization you work for.
To what extent do you agree or disagree with each of the following statements?
Given the chance, I would change my present type of work for something different.
Would you say...
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6334/vshow}{GSS Data Explorer page for \code{chngewrk}}.
}

\seealso{
Other Work Orientation: 
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
