% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_304.R
\name{chngtme}
\alias{chngtme}
\title{How often r allowed change schedule}
\source{
General Social Survey https://gss.norc.org
}
\description{
chngtme
}
\details{
Question How often are you allowed to change your starting and quitting times on a daily basis?
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2766/vshow}{GSS Data Explorer page for \code{chngtme}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{waypaid}},
\code{\link{wrksched}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Employment: 
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{waypaid}},
\code{\link{wrksched}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Job: 
\code{\link{wrksched}},
\code{\link{yearsjob}}
}
\concept{Employment}
\concept{Job}
\concept{Quality of Working Life}
\keyword{variable}
