% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{chngwork}
\alias{chngwork}
\title{I would change my job for something else}
\source{
General Social Survey https://gss.norc.org
}
\description{
chngwork
}
\details{
Question To what extent do you agree or disagree with each of the following statements?
C. Given the chance, I would change my present type of work for something different.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4587/vshow}{GSS Data Explorer page for \code{chngwork}}.
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
