% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{chores2}
\alias{chores2}
\title{Help for household tasks #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
chores2
}
\details{
Question First, there are some household and garden jobs you can''t really Do alone --for example, you may need someone to hold a ladder, or to help you move furniture.
B. Second Choice
}
\section{Values}{

\itemize{
\item \verb{[1]} spouse
\item \verb{[2]} mother
\item \verb{[3]} father
\item \verb{[4]} daughter
\item \verb{[5]} son
\item \verb{[6]} sister
\item \verb{[7]} brother
\item \verb{[8]} other relative
\item \verb{[9]} closest friend
\item \verb{[10]} other friend
\item \verb{[11]} neighbor
\item \verb{[12]} coworker
\item \verb{[13]} social service
\item \verb{[14]} paid help
\item \verb{[15]} other
\item \verb{[16]} no one
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4170/vshow}{GSS Data Explorer page for \code{chores2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab spouse \tab mother \tab father \tab daughter \tab son \tab sister \tab brother \tab other relative \tab closest friend \tab other friend \tab neighbor \tab coworker \tab social service \tab paid help \tab other \tab no one \tab no answer \tab Total \cr
   1986 \tab 42 \tab 49 \tab 78 \tab 67 \tab 116 \tab 218 \tab 41 \tab 92 \tab 101 \tab 186 \tab 106 \tab 218 \tab 27 \tab 1 \tab 47 \tab 6 \tab 46 \tab 29 \tab 1470 \cr
   Total \tab 42 \tab 49 \tab 78 \tab 67 \tab 116 \tab 218 \tab 41 \tab 92 \tab 101 \tab 186 \tab 106 \tab 218 \tab 27 \tab 1 \tab 47 \tab 6 \tab 46 \tab 29 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
