% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{chores2}
\alias{chores2}
\title{Help for household tasks #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
chores2
}
\details{
Question First, there are some household and garden jobs you can''t really Do alone --for example, you may need someone to hold a ladder, or to help you move furniture.
B. Second Choice
}
\section{Values}{

\itemize{
\item \code{1} spouse
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other relative
\item \code{9} closest friend
\item \code{10} other friend
\item \code{11} neighbor
\item \code{12} coworker
\item \code{13} social service
\item \code{14} paid help
\item \code{15} other
\item \code{16} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4170/vshow}{GSS Data Explorer page for \code{chores2}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}

Other Social Networks: 
\code{\link{chores1}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{Social Networks}
\keyword{variable}
