% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{churchtx}
\alias{churchtx}
\title{Membership in church-related groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
churchtx
}
\details{
Question Besides belonging to a congregation or religion, do you belong to any church-related groups such as a fellowship or study groups, a choir, board, or committee?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3108/vshow}{GSS Data Explorer page for \code{churchtx}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{infrmgrp}},
\code{\link{typfrat}},
\code{\link{typpolit}},
\code{\link{typserv}},
\code{\link{typvet}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
