% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{churhgrp}
\alias{churhgrp}
\title{Is church group church itself}
\source{
General Social Survey https://gss.norc.org
}
\description{
churhgrp
}
\details{
Question You said you were a member of a church affiliated group. Is that group or organization the church (synagogue) itself, or some other group related to the church?
}
\section{Values}{

\itemize{
\item \code{1} church
\item \code{2} related group
\item \code{3} both
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1012/vshow}{GSS Data Explorer page for \code{churhgrp}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{caregrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{mostact}},
\code{\link{servegrp}},
\code{\link{writegrp}}

Other Religion: 
\code{\link{actchurh}}

Other Voluntary Associations: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{caregrp}},
\code{\link{givegrp}},
\code{\link{leadgrp}},
\code{\link{mostact}},
\code{\link{servegrp}},
\code{\link{writegrp}}
}
\concept{Religion}
\concept{Socio-political participation}
\concept{Voluntary Associations}
\keyword{variable}
