% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{churhpow}
\alias{churhpow}
\title{Do churches have too much power?}
\source{
General Social Survey https://gss.norc.org
}
\description{
churhpow
}
\details{
Question Do you think that churches and religious organizations in this country have too much power or too little power?
}
\section{Values}{

\itemize{
\item \code{1} far too much power
\item \code{2} too much power
\item \code{3} about the right amount of power
\item \code{4} too little power
\item \code{5} far too little power
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4649/vshow}{GSS Data Explorer page for \code{churhpow}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}

Other Religion: 
\code{\link{afterlif}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
