% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{churhsch}
\alias{churhsch}
\title{Number of years r attended church school}
\source{
General Social Survey https://gss.norc.org
}
\description{
churhsch
}
\details{
Question Did you get any of your grade or high school education in parochial schools or other schools run by religious groups?
Yes................(ASK A)
No..............(GO TO  SUNSCH16)
A. How many years did you attend religious schools?
}
\section{Values}{

\itemize{
\item \code{0} never went or less than half year
\item \code{95} one or more year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1068/vshow}{GSS Data Explorer page for \code{churhsch}}.
}

\seealso{
Other Religion: 
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}

Other Religion: 
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
