% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cideage}
\alias{cideage}
\title{Age of murder victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cideage
}
\details{
Question We would like to know a few things about that person.
B. How old was that person?  Was (he/she) . . . (READ LIST)
}
\section{Values}{

\itemize{
\item \code{1} 10 years or under
\item \code{2} 11-20
\item \code{3} 21-40
\item \code{4} 41 year or older
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4993/vshow}{GSS Data Explorer page for \code{cideage}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 10 years or under \tab 11-20 \tab 21-40 \tab 41 year or older \tab no answer \tab Total \cr
   1988 \tab 1325 \tab 2 \tab 26 \tab 90 \tab 35 \tab 3 \tab 1481 \cr
   1989 \tab 1375 \tab 1 \tab 26 \tab 99 \tab 33 \tab 3 \tab 1537 \cr
   1990 \tab 1244 \tab 2 \tab 23 \tab 75 \tab 26 \tab 2 \tab 1372 \cr
   Total \tab 3944 \tab 5 \tab 75 \tab 264 \tab 94 \tab 8 \tab 4390 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
