% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{cideage3}
\alias{cideage3}
\title{Age of murder victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cideage3
}
\details{
Question We would like to know a few things about that person.
B. How old was that person?  Was (he/she) ... (READ LIST)
}
\section{Values}{

\itemize{
\item \code{1} 10 years or under
\item \code{2} 11-20
\item \code{3} 21-40
\item \code{4} 41 year or older
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5003/vshow}{GSS Data Explorer page for \code{cideage3}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsknow}},
\code{\link{aidswho}},
\code{\link{cideage2}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex3}},
\code{\link{cidewho3}}

Other Homicide: 
\code{\link{cideage2}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex3}},
\code{\link{cidewho3}}
}
\concept{Homicide}
\concept{Victims & AIDS}
\keyword{variable}
