% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{ciderac2}
\alias{ciderac2}
\title{Race of murder victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
ciderac2
}
\details{
Question We would like to know a few things about that person.
C. What was that person''s race?  Was it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \verb{[1]} black
\item \verb{[2]} white
\item \verb{[3]} hispanic
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4999/vshow}{GSS Data Explorer page for \code{ciderac2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab black \tab white \tab hispanic \tab no answer \tab other \tab Total \cr
   1989 \tab 1478 \tab 28 \tab 23 \tab 4 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 1325 \tab 29 \tab 13 \tab 1 \tab 2 \tab 2 \tab 1372 \cr
   Total \tab 2803 \tab 57 \tab 36 \tab 5 \tab 6 \tab 2 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsknow}},
\code{\link{aidswho}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{ciderac3}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex3}},
\code{\link{cidewho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
