% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{ciderac3}
\alias{ciderac3}
\title{Race of murder victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
ciderac3
}
\details{
Question We would like to know a few things about that person.
C. What was that person''s race?  Was it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \code{1} black
\item \code{2} white
\item \code{3} hispanic
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5004/vshow}{GSS Data Explorer page for \code{ciderac3}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab black \tab white \tab hispanic \tab no answer \tab Total \cr
   1989 \tab 1512 \tab 14 \tab 6 \tab 2 \tab 3 \tab 1537 \cr
   1990 \tab 1352 \tab 11 \tab 5 \tab 1 \tab 3 \tab 1372 \cr
   Total \tab 2864 \tab 25 \tab 11 \tab 3 \tab 6 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsknow}},
\code{\link{aidswho}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{ciderac2}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex3}},
\code{\link{cidewho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
