% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{ciderace}
\alias{ciderace}
\title{Race of murder victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
ciderace
}
\details{
Question We would like to know a few things about that person.
C. What was that person''s race?  Was it black, white, hispanic or other?
}
\section{Values}{

\itemize{
\item \code{1} black
\item \code{2} white
\item \code{3} hispanic
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4994/vshow}{GSS Data Explorer page for \code{ciderace}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab black \tab white \tab hispanic \tab other \tab no answer \tab Total \cr
   1988 \tab 1325 \tab 48 \tab 81 \tab 18 \tab 7 \tab 2 \tab 1481 \cr
   1989 \tab 1375 \tab 45 \tab 100 \tab 14 \tab - \tab 3 \tab 1537 \cr
   1990 \tab 1244 \tab 51 \tab 62 \tab 8 \tab 5 \tab 2 \tab 1372 \cr
   Total \tab 3944 \tab 144 \tab 243 \tab 40 \tab 12 \tab 7 \tab 4390 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{cidereg}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
