% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cidereg}
\alias{cidereg}
\title{Region of murder victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidereg
}
\details{
Question We would like to know a few things about that person.
D. What State did this person live in?
}
\section{Values}{

\itemize{
\item \code{0} foreign
\item \code{1} new england
\item \code{2} middle atlantic
\item \code{3} e. nor. central
\item \code{4} w. nor. central
\item \code{5} south atlantic
\item \code{6} e. sou. central
\item \code{7} w. sou. central
\item \code{8} mountain
\item \code{9} pacific
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4995/vshow}{GSS Data Explorer page for \code{cidereg}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab foreign \tab new england \tab middle atlantic \tab e. nor. central \tab w. nor. central \tab south atlantic \tab e. sou. central \tab w. sou. central \tab mountain \tab pacific \tab no answer \tab Total \cr
   1989 \tab 1375 \tab 1 \tab 4 \tab 20 \tab 28 \tab 7 \tab 32 \tab 12 \tab 22 \tab 12 \tab 18 \tab 6 \tab 1537 \cr
   1990 \tab 1244 \tab 1 \tab 3 \tab 18 \tab 24 \tab 6 \tab 26 \tab 10 \tab 13 \tab 14 \tab 11 \tab 2 \tab 1372 \cr
   Total \tab 2619 \tab 2 \tab 7 \tab 38 \tab 52 \tab 13 \tab 58 \tab 22 \tab 35 \tab 26 \tab 29 \tab 8 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
