% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cidereg}
\alias{cidereg}
\title{Region of murder victim closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidereg
}
\details{
Question We would like to know a few things about that person.
D. What State did this person live in?
}
\section{Values}{

\itemize{
\item \code{0} foreign
\item \code{1} new england
\item \code{2} middle atlantic
\item \code{3} e. nor. central
\item \code{4} w. nor. central
\item \code{5} south atlantic
\item \code{6} e. sou. central
\item \code{7} w. sou. central
\item \code{8} mountain
\item \code{9} pacific
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4995/vshow}{GSS Data Explorer page for \code{cidereg}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}}

Other Homicide: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Homicide}
\concept{Victims & AIDS}
\keyword{variable}
