% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cidesex2}
\alias{cidesex2}
\title{Sex of murder victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidesex2
}
\details{
Question We would like to know a few things about that person.
A. Was that person male or female?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4997/vshow}{GSS Data Explorer page for \code{cidesex2}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidesex}},
\code{\link{cidewho}},
\code{\link{cidewho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
