% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{cidesex3}
\alias{cidesex3}
\title{Sex of murder victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidesex3
}
\details{
Question We would like to know a few things about that person.
A. Was that person male or female?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5002/vshow}{GSS Data Explorer page for \code{cidesex3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1989 \tab 1512 \tab 19 \tab 3 \tab 3 \tab 1537 \cr
   1990 \tab 1352 \tab 13 \tab 5 \tab 2 \tab 1372 \cr
   Total \tab 2864 \tab 32 \tab 8 \tab 5 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsknow}},
\code{\link{aidswho}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidewho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
