% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cidewho}
\alias{cidewho}
\title{R's relationship to closest murderee}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidewho
}
\details{
Question Think about the person you know best who was a victim of homicide. Please tell me the letter of the category on the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \code{1} husband or wife
\item \code{2} partner or lover
\item \code{3} son or daughter
\item \code{4} other relative
\item \code{5} friend
\item \code{6} neighbor
\item \code{7} co-worker
\item \code{8} acquaintance
\item \code{9} patient
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4991/vshow}{GSS Data Explorer page for \code{cidewho}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab partner or lover \tab other relative \tab friend \tab neighbor \tab co-worker \tab acquaintance \tab other \tab no answer \tab son or daughter \tab patient \tab Total \cr
   1988 \tab 1325 \tab 2 \tab 17 \tab 70 \tab 10 \tab 13 \tab 37 \tab 5 \tab 2 \tab - \tab - \tab 1481 \cr
   1989 \tab 1375 \tab - \tab 19 \tab 60 \tab 10 \tab 9 \tab 52 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1537 \cr
   1990 \tab 1244 \tab 1 \tab 15 \tab 45 \tab 5 \tab 9 \tab 43 \tab 4 \tab 2 \tab 1 \tab 3 \tab 1372 \cr
   Total \tab 3944 \tab 3 \tab 51 \tab 175 \tab 25 \tab 31 \tab 132 \tab 16 \tab 7 \tab 2 \tab 4 \tab 4390 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho2}}
}
\concept{Victims & AIDS}
\keyword{variable}
