% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_546.R
\name{cidewho2}
\alias{cidewho2}
\title{R's relationship to 2nd closest murderee}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidewho2
}
\details{
Question Think about the person you knew next best who was a victim of homicide. Please tell me the letter of the category On the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \code{1} husband or wife
\item \code{2} partner or lover
\item \code{3} son or daughter
\item \code{4} other relative
\item \code{5} friend
\item \code{6} neighbor
\item \code{7} co-worker
\item \code{8} acquaintance
\item \code{9} patient
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4996/vshow}{GSS Data Explorer page for \code{cidewho2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab other relative \tab friend \tab neighbor \tab co-worker \tab acquaintance \tab other \tab no answer \tab partner or lover \tab Total \cr
   1989 \tab 1478 \tab 3 \tab 19 \tab 5 \tab 6 \tab 20 \tab 2 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 1325 \tab 3 \tab 18 \tab 1 \tab 1 \tab 20 \tab 1 \tab 2 \tab 1 \tab 1372 \cr
   Total \tab 2803 \tab 6 \tab 37 \tab 6 \tab 7 \tab 40 \tab 3 \tab 6 \tab 1 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{cideage}},
\code{\link{cideknew}},
\code{\link{ciderace}},
\code{\link{cidereg}},
\code{\link{cidesex}},
\code{\link{cidesex2}},
\code{\link{cidewho}}
}
\concept{Victims & AIDS}
\keyword{variable}
