% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_547.R
\name{cidewho3}
\alias{cidewho3}
\title{R's relationship to 3rd closest murderee}
\source{
General Social Survey https://gss.norc.org
}
\description{
cidewho3
}
\details{
Question Think about the person you knew next best who was a victim of homicide. Please tell me the letter of the category On the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \verb{[1]} husband or wife
\item \verb{[2]} partner or lover
\item \verb{[3]} son or daughter
\item \verb{[4]} other relative
\item \verb{[5]} friend
\item \verb{[6]} neighbor
\item \verb{[7]} co-worker
\item \verb{[8]} acquaintance
\item \verb{[9]} patient
\item \verb{[10]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5001/vshow}{GSS Data Explorer page for \code{cidewho3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab other relative \tab friend \tab co-worker \tab acquaintance \tab patient \tab no answer \tab neighbor \tab Total \cr
   1989 \tab 1512 \tab 1 \tab 9 \tab 1 \tab 10 \tab 1 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 1352 \tab 2 \tab 6 \tab 2 \tab 6 \tab - \tab 2 \tab 2 \tab 1372 \cr
   Total \tab 2864 \tab 3 \tab 15 \tab 3 \tab 16 \tab 1 \tab 5 \tab 2 \tab 2909 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsknow}},
\code{\link{aidswho}},
\code{\link{cideage2}},
\code{\link{cideage3}},
\code{\link{ciderac2}},
\code{\link{ciderac3}},
\code{\link{cidereg2}},
\code{\link{cidereg3}},
\code{\link{cidesex3}}
}
\concept{Victims & AIDS}
\keyword{variable}
