% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{cigweek}
\alias{cigweek}
\title{Has r smoked cigarettes in last week}
\source{
General Social Survey https://gss.norc.org
}
\description{
cigweek
}
\details{
Question Do you smoke?
C. Have you, yourself, smoked any cigarettes in the past week?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/535/vshow}{GSS Data Explorer page for \code{cigweek}}.
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}

Other Smoking: 
\code{\link{evsmoke}},
\code{\link{quitsmk}},
\code{\link{smokecig}}
}
\concept{Core}
\concept{Smoking}
\keyword{variable}
