% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{citizen}
\alias{citizen}
\title{Are you a citizen of america?}
\source{
General Social Survey https://gss.norc.org
}
\description{
citizen
}
\details{
Question A. Are you a citizen of America?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4873/vshow}{GSS Data Explorer page for \code{citizen}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 1285 \tab 57 \tab 25 \tab 2904 \cr
   2004 \tab 1596 \tab 1168 \tab 47 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 1186 \tab 84 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 3639 \tab 188 \tab 30 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ethclose}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{parcit}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}}
}
\concept{ISSP National Identity}
\keyword{variable}
