% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{citviews}
\alias{citviews}
\title{How important: politician consider views of citizen}
\source{
General Social Survey https://gss.norc.org
}
\description{
citviews
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
D. That politicians take into account the views of citizens before making decisions
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4936/vshow}{GSS Data Explorer page for \code{citviews}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 14 \tab 6 \tab 18 \tab 54 \tab 98 \tab 177 \tab 1093 \tab 8 \tab 4 \tab 2812 \cr
   Total \tab 1340 \tab 14 \tab 6 \tab 18 \tab 54 \tab 98 \tab 177 \tab 1093 \tab 8 \tab 4 \tab 2812 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{eqtreat}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{polopts}},
\code{\link{rghtsmin}},
\code{\link{solok}}
}
\concept{ISSP Citizenship}
\keyword{variable}
