% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{civic}
\alias{civic}
\title{Does r follow public affairs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
civic
}
\details{
Question Some people seem to follow what's going on in government and public affairs most of the time, whether there's an election going on or not. Others aren't that interested. Would you say you follow what's going on in government and public affairs most of the time, some of the time, only now and then, or hardly at all?
}
\section{Values}{

\itemize{
\item \code{1} most of the time
\item \code{2} some of the time
\item \code{3} now and then
\item \code{4} hardly at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/773/vshow}{GSS Data Explorer page for \code{civic}}.
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\concept{Political}
\keyword{variable}
