% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{class1}
\alias{class1}
\title{Subjective class identification (2021)}
\source{
General Social Survey https://gss.norc.org
}
\description{
class1
}
\details{
Question Most people see themselves as belonging to a particular class. Please tell me which social class you would say you belong to?
}
\section{Values}{

\itemize{
\item \code{1} lower class
\item \code{2} working class
\item \code{3} lower middle class
\item \code{4} middle class
\item \code{5} upper middle class
\item \code{6} upper class
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7520/vshow}{GSS Data Explorer page for \code{class1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab lower class \tab working class \tab lower middle class \tab middle class \tab upper middle class \tab upper class \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 95 \tab 461 \tab 336 \tab 666 \tab 226 \tab 22 \tab 4 \tab 178 \tab 15 \tab 4032 \cr
   Total \tab 2029 \tab 95 \tab 461 \tab 336 \tab 666 \tab 226 \tab 22 \tab 4 \tab 178 \tab 15 \tab 4032 \cr
}
}

\keyword{variable}
