% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{classy}
\alias{classy}
\title{Subjective social class-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
classy
}
\details{
Question B. If you were asked to use one of five names for your social class, which would you say you belong in: the lower class, the working class, the middle class, the upper middle class, or the upper class?
}
\section{Values}{

\itemize{
\item \code{1} lower class
\item \code{2} working class
\item \code{3} middle class
\item \code{4} upper middle class
\item \code{5} upper class
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/569/vshow}{GSS Data Explorer page for \code{classy}}.
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()},
\code{\link{rank10}}

Other Social Class: 
\code{\link{class}()},
\code{\link{rank}()},
\code{\link{rank10}}

Other Split Ballots: 
\code{\link{class}()},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()}
}
\concept{Core}
\concept{Social Class}
\concept{Split Ballots}
\keyword{variable}
