% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{clerggov}
\alias{clerggov}
\title{Clergy shouldn't try to influence govt decisions}
\source{
General Social Survey https://gss.norc.org
}
\description{
clerggov
}
\details{
Question How much do you agree or disagree with each of the following…
D. Religious leaders should not try to influence government decisions.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4648/vshow}{GSS Data Explorer page for \code{clerggov}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 287 \tab 401 \tab 308 \tab 254 \tab 36 \tab 51 \tab 22 \tab 1517 \cr
   1998 \tab 1548 \tab 366 \tab 354 \tab 224 \tab 197 \tab 68 \tab 33 \tab 42 \tab 2832 \cr
   2008 \tab 658 \tab 515 \tab 401 \tab 206 \tab 182 \tab 48 \tab 12 \tab 1 \tab 2023 \cr
   Total \tab 2364 \tab 1168 \tab 1156 \tab 738 \tab 633 \tab 152 \tab 96 \tab 65 \tab 6372 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
