% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{clmtwrld}
\alias{clmtwrld}
\title{Impact of climate change for world}
\source{
General Social Survey https://gss.norc.org
}
\description{
clmtwrld
}
\details{
Question On a scale from 0 to 10, how bad or good do you think the impacts of climate change will be for the world as a whole?
}
\section{Values}{

\itemize{
\item \code{0} 0 - extremely bad
\item \code{10} 10 - extremely good
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7487/vshow}{GSS Data Explorer page for \code{clmtwrld}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 0 - extremely bad \tab 1 \tab 10 - extremely good \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab Total \cr
   2021 \tab 2003 \tab 428 \tab 137 \tab 25 \tab 243 \tab 239 \tab 184 \tab 346 \tab 56 \tab 44 \tab 20 \tab 15 \tab 83 \tab 209 \tab 4032 \cr
   Total \tab 2003 \tab 428 \tab 137 \tab 25 \tab 243 \tab 239 \tab 184 \tab 346 \tab 56 \tab 44 \tab 20 \tab 15 \tab 83 \tab 209 \tab 4032 \cr
}
}

\keyword{variable}
