% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{clmtwrld}
\alias{clmtwrld}
\title{Impact of climate change for world}
\source{
General Social Survey https://gss.norc.org
}
\description{
clmtwrld
}
\details{
Question On a scale from 0 to 10, how bad or good do you think the impacts of climate change will be for the world as a whole?
}
\section{Values}{

\itemize{
\item \code{0} 0 - extremely bad
\item \code{10} 10 - extremely good
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7487/vshow}{GSS Data Explorer page for \code{clmtwrld}}.
}

\keyword{variable}
