% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{cloning}
\alias{cloning}
\title{Cloning makes identical copy}
\source{
General Social Survey https://gss.norc.org
}
\description{
cloning
}
\details{
Question The cloning of living things produces genetically identical copies.
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3499/vshow}{GSS Data Explorer page for \code{cloning}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab can't choose \tab no answer \tab Total \cr
   2010 \tab 1103 \tab 751 \tab 82 \tab 104 \tab 4 \tab 2044 \cr
   Total \tab 1103 \tab 751 \tab 82 \tab 104 \tab 4 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
