% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_344.R
\name{closrgod}
\alias{closrgod}
\title{Desire to be closer to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
closrgod
}
\details{
Question I desire to be closer to God or in union with Him.
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never or almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3092/vshow}{GSS Data Explorer page for \code{closrgod}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never or almost never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 235 \tab 444 \tab 189 \tab 161 \tab 139 \tab 144 \tab 11 \tab 17 \tab 2812 \cr
   Total \tab 1472 \tab 235 \tab 444 \tab 189 \tab 161 \tab 139 \tab 144 \tab 11 \tab 17 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{blessngs}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{joylifts}},
\code{\link{relsprt1}},
\code{\link{relsprt2}}
}
\concept{Spirituality and Religion}
\keyword{variable}
