% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{clsenei}
\alias{clsenei}
\title{How close do you feel to your neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
clsenei
}
\details{
Question How close do you feel to…
\enumerate{
\item Your neighborhood (or village)
}
}
\section{Values}{

\itemize{
\item \code{1} very close
\item \code{2} close
\item \code{3} not very close
\item \code{4} not close at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4816/vshow}{GSS Data Explorer page for \code{clsenei}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very close \tab close \tab not very close \tab not close at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 197 \tab 556 \tab 405 \tab 168 \tab 27 \tab 14 \tab 2904 \cr
   Total \tab 1537 \tab 197 \tab 556 \tab 405 \tab 168 \tab 27 \tab 14 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{clsetown}}
}
\concept{ISSP National Identity}
\keyword{variable}
