% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{clsenei}
\alias{clsenei}
\title{How close do you feel to your neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
clsenei
}
\details{
Question How close do you feel to…
\enumerate{
\item Your neighborhood (or village)
}
}
\section{Values}{

\itemize{
\item \code{1} very close
\item \code{2} close
\item \code{3} not very close
\item \code{4} not close at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4816/vshow}{GSS Data Explorer page for \code{clsenei}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{clsetown}}
}
\concept{ISSP National Identity}
\keyword{variable}
