% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{clsenoam}
\alias{clsenoam}
\title{How close do you feel to north america}
\source{
General Social Survey https://gss.norc.org
}
\description{
clsenoam
}
\details{
Question How close do you feel to…
5. North America
}
\section{Values}{

\itemize{
\item \verb{[1]} very close
\item \verb{[2]} close
\item \verb{[3]} not very close
\item \verb{[4]} not close at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4820/vshow}{GSS Data Explorer page for \code{clsenoam}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very close \tab close \tab not very close \tab not close at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 226 \tab 502 \tab 367 \tab 144 \tab 86 \tab 42 \tab 2904 \cr
   2004 \tab 1596 \tab 317 \tab 461 \tab 292 \tab 89 \tab 56 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 273 \tab 507 \tab 288 \tab 139 \tab 64 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 816 \tab 1470 \tab 947 \tab 372 \tab 206 \tab 46 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{onenatn}}
}
\concept{ISSP National Identity}
\keyword{variable}
