% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{clsrltv}
\alias{clsrltv}
\title{Who are close relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
clsrltv
}
\details{
Question People may have different perceptions about who they consider to be close relatives. Imagine a person living with a partner and their children. Which of the following options comes closest to who you would consider to be his/her close relatives?
}
\section{Values}{

\itemize{
\item \code{1} children and partner only
\item \code{2} children, partner, parents and siblings
\item \code{3} children, partner, parents, siblings, cousins, aunts/uncles, nephews/nieces
\item \code{4} children, partner, parents, siblings, cousins, aunts/uncles, nephews/nieces, but others as well
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7777/vshow}{GSS Data Explorer page for \code{clsrltv}}.
}

\keyword{variable}
