% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{cntctfrd}
\alias{cntctfrd}
\title{How often r contacts their most-contacted close friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
cntctfrd
}
\details{
Question And, think about the close friend you have contact with most frequently: How often do you have contact with that close friend,
either face-to-face, by phone, internet or any other communication device?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several times a week
\item \verb{[3]} once a week
\item \verb{[4]} two to three times a month
\item \verb{[5]} once a month
\item \verb{[6]} several times a year
\item \verb{[7]} less often
\item \verb{[8]} never
\item \verb{[9]} i do not have any close friends
\item \verb{[10]} the close friend i have contact with the most frequently lives with me
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7269/vshow}{GSS Data Explorer page for \code{cntctfrd}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab daily \tab several times a week \tab once a week \tab two to three times a month \tab once a month \tab several times a year \tab less often \tab never \tab i do not have any close friends \tab the close friend i have contact with the most frequently lives with me \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 274 \tab 402 \tab 171 \tab 137 \tab 54 \tab 62 \tab 22 \tab 23 \tab 21 \tab 2 \tab 4 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 274 \tab 402 \tab 171 \tab 137 \tab 54 \tab 62 \tab 22 \tab 23 \tab 21 \tab 2 \tab 4 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
