% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{cntctkid}
\alias{cntctkid}
\title{How often r contacts their most-contacted adult child}
\source{
General Social Survey https://gss.norc.org
}
\description{
cntctkid
}
\details{
Question Think about your adult child you have contact with most frequently: How often do you have contact with this child aged at least 18, either face-to-face, by phone, internet or any other communication device?
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} several times a week
\item \code{3} once a week
\item \code{4} two to three times a month
\item \code{5} once a month
\item \code{6} several times a year
\item \code{7} less often
\item \code{8} never
\item \code{9} i do not have any adult children
\item \code{10} the adult child i have contact with most frequently lives with me
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7267/vshow}{GSS Data Explorer page for \code{cntctkid}}.
}

\keyword{variable}
