% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{cntrlife}
\alias{cntrlife}
\title{Control r feels over own life}
\source{
General Social Survey https://gss.norc.org
}
\description{
cntrlife
}
\details{
Question Some people feel they have completely free choice and control over the way their lives turn out, and other people feel that what they themselves do has no real effect on what happens to them. Please use this scale to indicate how much freedom of choice and control you feel you have over the way your life turns out.
}
\section{Values}{

\itemize{
\item \code{1} no choice and control
\item \code{4} some choice and control
\item \code{7} a great deal of choice and control
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2676/vshow}{GSS Data Explorer page for \code{cntrlife}}.
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{expunpop}},
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{freenow}},
\code{\link{freeprss}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}},
\code{\link{wlthpov}}
}
\concept{Freedom}
\keyword{variable}
