% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{cobarate}
\alias{cobarate}
\title{Rate of ba degree earning at college partner attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
cobarate
}
\details{
Question In what year did your partner receive the (COLLEGE NAME) degree?
K. Graduation Rate
}
\section{Values}{

\itemize{
\item \code{1} 0-39.9pct
\item \code{2} 40-57.9pct
\item \code{3} 58pct or over
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6164/vshow}{GSS Data Explorer page for \code{cobarate}}.
}

\seealso{
Other Core: 
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
