% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{coden}
\alias{coden}
\title{Partner's denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
coden
}
\details{
Question What is your (PARTNER'S) religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion? A. What specific denomination is that, if any? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \code{10} american baptist association
\item \code{11} american baptist churches in the u.s.a.
\item \code{12} national baptist convention of america
\item \code{13} national baptist convention u.s.a., inc.
\item \code{14} southern baptist convention
\item \code{15} other baptist churches
\item \code{18} baptist, don't know which
\item \code{20} african methodist episcopal church
\item \code{21} african methodist episcopal zion church
\item \code{22} united methodist church
\item \code{23} other methodist churches
\item \code{28} methodist, don't know which
\item \code{30} american lutheran church
\item \code{31} lutheran church in america
\item \code{32} lutheran church - missouri synod
\item \code{33} wisconsin evangelical lutheran synod
\item \code{34} other lutheran churches
\item \code{35} evangelical lutheran
\item \code{38} lutheran, don't know which
\item \code{40} presbyterian church in the u.s.
\item \code{41} united presbyterian church in the u.s.a.
\item \code{42} other presbyterian churches
\item \code{43} presbyterian, merged
\item \code{48} presbyterian, don't know which
\item \code{50} episcopal church
\item \code{60} other, specify
\item \code{70} no denomination or non-denominational church
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6170/vshow}{GSS Data Explorer page for \code{coden}}.

Counts by year:\tabular{llllllllllllllllllllllllllll}{
   year \tab iap \tab american baptist association \tab american baptist churches in the u.s.a. \tab national baptist convention of america \tab southern baptist convention \tab baptist, don't know which \tab united methodist church \tab methodist, don't know which \tab american lutheran church \tab lutheran church in america \tab evangelical lutheran \tab united presbyterian church in the u.s.a. \tab presbyterian, don't know which \tab episcopal church \tab other, specify \tab no denomination or non-denominational church \tab don't know \tab lutheran, don't know which \tab no answer \tab other baptist churches \tab african methodist episcopal church \tab national baptist convention u.s.a., inc. \tab lutheran church - missouri synod \tab presbyterian church in the u.s. \tab presbyterian, merged \tab other lutheran churches \tab Total \cr
   2012 \tab 1920 \tab 1 \tab 1 \tab 1 \tab 11 \tab 7 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 14 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2471 \tab - \tab 2 \tab - \tab 8 \tab 15 \tab 6 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 10 \tab 15 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2766 \tab 1 \tab 2 \tab 1 \tab 11 \tab 14 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 6 \tab 14 \tab 23 \tab 5 \tab 7 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2272 \tab 3 \tab - \tab - \tab 11 \tab 10 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 15 \tab 18 \tab 4 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3942 \tab 2 \tab 1 \tab 2 \tab 6 \tab 24 \tab 5 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 6 \tab 21 \tab 8 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4032 \cr
   2022 \tab 3456 \tab 5 \tab 1 \tab - \tab 6 \tab 24 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 10 \tab 26 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3544 \cr
   Total \tab 16827 \tab 12 \tab 7 \tab 4 \tab 53 \tab 94 \tab 20 \tab 8 \tab 6 \tab 3 \tab 3 \tab 3 \tab 10 \tab 16 \tab 58 \tab 117 \tab 26 \tab 19 \tab 6 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
