% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{codipged}
\alias{codipged}
\title{Partner diploma, ged, or other}
\source{
General Social Survey https://gss.norc.org
}
\description{
codipged
}
\details{
Question Was that a diploma awarded by (RESPONDENT'S PARTNER)''s high school at graduation at the end of 12th grade, a GED awarded after (RESPONDENT''S PARTNER) took a test, or something else?
}
\section{Values}{

\itemize{
\item \verb{[1]} high school diploma
\item \verb{[2]} ged
\item \verb{[3]} other
\item \verb{[5]} high school diploma after post high school class
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6162/vshow}{GSS Data Explorer page for \code{codipged}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab high school diploma \tab ged \tab other \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1846 \tab 111 \tab 17 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2360 \tab 146 \tab 26 \tab 2 \tab 1 \tab 3 \tab 2538 \cr
   2016 \tab 2644 \tab 196 \tab 23 \tab 3 \tab 1 \tab - \tab 2867 \cr
   Total \tab 6850 \tab 453 \tab 66 \tab 5 \tab 2 \tab 3 \tab 7379 \cr
}
}

\seealso{
Other Core: 
\code{\link{cosector}},
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\keyword{variable}
