% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{coeduc}
\alias{coeduc}
\title{Highest year school completed, partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
coeduc
}
\details{
Question RESPONDENT'S PARTNER'S EDUCATION
}
\section{Values}{

\itemize{
\item \code{0} no formal schooling
\item \code{1} 1st grade
\item \code{2} 2nd grade
\item \code{3} 3rd grade
\item \code{4} 4th grade
\item \code{5} 5th grade
\item \code{6} 6th grade
\item \code{7} 7th grade
\item \code{8} 8th grade
\item \code{9} 9th grade
\item \code{10} 10th grade
\item \code{11} 11th grade
\item \code{12} 12th grade
\item \code{13} 1 year of college
\item \code{14} 2 years of college
\item \code{15} 3 years of college
\item \code{16} 4 years of college
\item \code{17} 5 years of college
\item \code{18} 6 years of college
\item \code{19} 7 years of college
\item \code{20} 8 or more years of college
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6160/vshow}{GSS Data Explorer page for \code{coeduc}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 4 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab 3 \tab no formal schooling \tab 5 \tab Total \cr
   2012 \tab 1806 \tab 11 \tab 13 \tab 55 \tab 8 \tab 16 \tab 4 \tab 22 \tab 3 \tab 4 \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 4 \tab 14 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2325 \tab 6 \tab 13 \tab 88 \tab 15 \tab 18 \tab 10 \tab 21 \tab 4 \tab 6 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 14 \tab 4 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2603 \tab 9 \tab 16 \tab 100 \tab 24 \tab 22 \tab 8 \tab 30 \tab 10 \tab 4 \tab 4 \tab 8 \tab 1 \tab 4 \tab 1 \tab 4 \tab 8 \tab 7 \tab 3 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 2117 \tab 10 \tab 14 \tab 80 \tab 16 \tab 32 \tab 9 \tab 38 \tab 6 \tab 7 \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 2348 \cr
   2021 \tab 3660 \tab 7 \tab 11 \tab 119 \tab 22 \tab 46 \tab 15 \tab 69 \tab 7 \tab 15 \tab 4 \tab 10 \tab 3 \tab 2 \tab 2 \tab 3 \tab 6 \tab 11 \tab 18 \tab 1 \tab 1 \tab - \tab 4032 \cr
   2022 \tab 3228 \tab 10 \tab 10 \tab 104 \tab 21 \tab 38 \tab 17 \tab 58 \tab 11 \tab 12 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 8 \tab - \tab 11 \tab 1 \tab 2 \tab 2 \tab 3544 \cr
   Total \tab 15739 \tab 53 \tab 77 \tab 546 \tab 106 \tab 172 \tab 63 \tab 238 \tab 41 \tab 48 \tab 16 \tab 25 \tab 6 \tab 15 \tab 10 \tab 17 \tab 33 \tab 51 \tab 38 \tab 3 \tab 4 \tab 2 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{age}},
\code{\link{agekdbrn}},
\code{\link{codeg}},
\code{\link{degree}},
\code{\link{educ}},
\code{\link{maeduc}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{speduc}}
}
\concept{Core}
\keyword{variable}
