% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{coeftotlt}
\alias{coeftotlt}
\title{Total number of students enrolled at college partner attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
coeftotlt
}
\details{
Question In what year did your partner receive the (COLLEGE NAME) degree?
J. Total Fall Enrollment
}
\section{Values}{

\itemize{
\item \code{1} 0-467 students enrolled
\item \code{2} 468 or more students
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6166/vshow}{GSS Data Explorer page for \code{coeftotlt}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 468 or more students \tab 0-467 students enrolled \tab no answer \tab Total \cr
   2012 \tab 1939 \tab 35 \tab - \tab - \tab 1974 \cr
   2014 \tab 2507 \tab 29 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2806 \tab 60 \tab - \tab 1 \tab 2867 \cr
   Total \tab 7252 \tab 124 \tab 2 \tab 1 \tab 7379 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
