% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_04.R
\name{coevwork}
\alias{coevwork}
\title{Partner ever work as long as a year}
\source{
General Social Survey https://gss.norc.org
}
\description{
coevwork
}
\details{
Question Last week was your partner working full time, part time, going to school, keeping house, or what? C. IF RETIRED, IN SCHOOL, KEEPING HOUSE, OR OTHER: Did (he/she) ever work for as long as one year?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6154/vshow}{GSS Data Explorer page for \code{coevwork}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 1938 \tab 26 \tab 8 \tab 1 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 2484 \tab 35 \tab 14 \tab 2 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 2795 \tab 53 \tab 16 \tab 3 \tab - \tab - \tab 2867 \cr
   2018 \tab 2300 \tab 38 \tab 10 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3941 \tab 67 \tab 7 \tab 9 \tab 8 \tab - \tab 4032 \cr
   2022 \tab 3473 \tab 53 \tab 12 \tab 2 \tab 3 \tab 1 \tab 3544 \cr
   Total \tab 16931 \tab 272 \tab 67 \tab 17 \tab 15 \tab 1 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{cohrs1}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{divorce}},
\code{\link{spevwork}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spwrksta}},
\code{\link{widowed}}
}
\concept{Core}
\keyword{variable}
