% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{cohort}
\alias{cohort}
\title{Year of birth}
\source{
General Social Survey https://gss.norc.org
}
\description{
cohort
}
\details{
Question Birth cohort of respondent.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5507/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 1883 \tab 1884 \tab 1885 \tab 1887 \tab 1888 \tab 1889 \tab 1890 \tab 1891 \tab 1892 \tab 1893 \tab 1894 \tab 1895 \tab 1896 \tab 1897 \tab 1898 \tab 1899 \tab 1900 \tab 1901 \tab 1902 \tab 1903 \tab 1904 \tab 1905 \tab 1906 \tab 1907 \tab 1908 \tab 1909 \tab 1910 \tab 1911 \tab 1912 \tab 1913 \tab 1914 \tab 1915 \tab 1916 \tab 1917 \tab 1918 \tab 1919 \tab 1920 \tab 1921 \tab 1922 \tab 1923 \tab 1924 \tab 1925 \tab 1926 \tab 1927 \tab 1928 \tab 1929 \tab 1930 \tab 1931 \tab 1932 \tab 1933 \tab 1934 \tab 1935 \tab 1936 \tab 1937 \tab 1938 \tab 1939 \tab 1940 \tab 1941 \tab 1942 \tab 1943 \tab 1944 \tab 1945 \tab 1946 \tab 1947 \tab 1948 \tab 1949 \tab 1950 \tab 1951 \tab 1952 \tab 1953 \tab 1954 \tab no answer \tab 1955 \tab 1886 \tab 1956 \tab 1957 \tab 1958 \tab 1959 \tab 1960 \tab 1961 \tab 1962 \tab 1963 \tab 1964 \tab 1965 \tab 1966 \tab 1967 \tab 1968 \tab 1969 \tab 1970 \tab 1971 \tab 1972 \tab 1973 \tab 1974 \tab 1975 \tab 1976 \tab 1977 \tab 1978 \tab 1979 \tab 1980 \tab 1981 \tab 1982 \tab 1983 \tab 1984 \tab 1985 \tab 1986 \tab 1987 \tab 1988 \tab 1989 \tab 1990 \tab 1991 \tab 1992 \tab 1993 \tab 1994 \tab 1995 \tab 1996 \tab 1997 \tab 1998 \tab 1999 \tab 2000 \tab 2001 \tab 2002 \tab 2003 \tab 2004 \tab 2005 \tab 2006 \tab Total \cr
   1972 \tab 2 \tab 2 \tab 2 \tab 3 \tab 3 \tab 4 \tab 6 \tab 5 \tab 6 \tab 4 \tab 12 \tab 10 \tab 12 \tab 13 \tab 17 \tab 12 \tab 19 \tab 9 \tab 17 \tab 16 \tab 20 \tab 13 \tab 11 \tab 24 \tab 27 \tab 23 \tab 15 \tab 21 \tab 34 \tab 25 \tab 24 \tab 26 \tab 31 \tab 26 \tab 24 \tab 29 \tab 32 \tab 26 \tab 28 \tab 36 \tab 38 \tab 30 \tab 31 \tab 42 \tab 23 \tab 31 \tab 29 \tab 22 \tab 16 \tab 21 \tab 23 \tab 18 \tab 28 \tab 29 \tab 32 \tab 26 \tab 25 \tab 25 \tab 50 \tab 45 \tab 30 \tab 37 \tab 48 \tab 40 \tab 31 \tab 31 \tab 38 \tab 37 \tab 34 \tab 24 \tab 5 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 8 \tab 6 \tab 10 \tab 7 \tab 7 \tab 25 \tab 11 \tab 12 \tab 14 \tab 17 \tab 15 \tab 17 \tab 16 \tab 24 \tab 17 \tab 24 \tab 20 \tab 23 \tab 16 \tab 21 \tab 28 \tab 29 \tab 19 \tab 33 \tab 18 \tab 30 \tab 20 \tab 29 \tab 28 \tab 20 \tab 30 \tab 24 \tab 23 \tab 32 \tab 22 \tab 21 \tab 24 \tab 24 \tab 23 \tab 24 \tab 27 \tab 32 \tab 28 \tab 41 \tab 31 \tab 35 \tab 21 \tab 22 \tab 29 \tab 29 \tab 34 \tab 30 \tab 40 \tab 33 \tab 32 \tab 43 \tab 33 \tab 42 \tab 32 \tab 29 \tab 29 \tab 4 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab 5 \tab 1 \tab 3 \tab 2 \tab 2 \tab 6 \tab 1 \tab 6 \tab 8 \tab 6 \tab 8 \tab 5 \tab 11 \tab 8 \tab 13 \tab 12 \tab 21 \tab 17 \tab 10 \tab 27 \tab 17 \tab 19 \tab 18 \tab 25 \tab 11 \tab 17 \tab 10 \tab 22 \tab 23 \tab 25 \tab 26 \tab 19 \tab 24 \tab 19 \tab 26 \tab 21 \tab 27 \tab 32 \tab 23 \tab 15 \tab 28 \tab 20 \tab 25 \tab 22 \tab 28 \tab 24 \tab 32 \tab 15 \tab 26 \tab 31 \tab 35 \tab 26 \tab 19 \tab 26 \tab 24 \tab 22 \tab 30 \tab 38 \tab 35 \tab 30 \tab 36 \tab 43 \tab 41 \tab 28 \tab 34 \tab 50 \tab 29 \tab 33 \tab 25 \tab 6 \tab 21 \tab 1 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab 3 \tab 4 \tab 4 \tab 9 \tab 7 \tab 7 \tab 6 \tab 13 \tab 12 \tab 3 \tab 12 \tab 10 \tab 11 \tab 8 \tab 28 \tab 21 \tab 16 \tab 21 \tab 25 \tab 19 \tab 18 \tab 18 \tab 26 \tab 13 \tab 19 \tab 21 \tab 24 \tab 20 \tab 16 \tab 21 \tab 24 \tab 21 \tab 26 \tab 19 \tab 31 \tab 21 \tab 23 \tab 14 \tab 31 \tab 26 \tab 18 \tab 33 \tab 21 \tab 28 \tab 33 \tab 21 \tab 30 \tab 17 \tab 25 \tab 24 \tab 26 \tab 35 \tab 37 \tab 19 \tab 30 \tab 34 \tab 36 \tab 38 \tab 36 \tab 46 \tab 42 \tab 35 \tab 31 \tab 26 \tab 5 \tab 33 \tab 3 \tab 38 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab - \tab 7 \tab 1 \tab 5 \tab 6 \tab 4 \tab 7 \tab 4 \tab 4 \tab 5 \tab 9 \tab 10 \tab 13 \tab 10 \tab 17 \tab 10 \tab 9 \tab 16 \tab 16 \tab 11 \tab 20 \tab 13 \tab 11 \tab 26 \tab 17 \tab 16 \tab 24 \tab 16 \tab 29 \tab 27 \tab 22 \tab 19 \tab 29 \tab 18 \tab 22 \tab 23 \tab 18 \tab 32 \tab 27 \tab 21 \tab 13 \tab 20 \tab 20 \tab 19 \tab 22 \tab 22 \tab 15 \tab 16 \tab 20 \tab 27 \tab 18 \tab 28 \tab 26 \tab 25 \tab 34 \tab 23 \tab 37 \tab 27 \tab 40 \tab 24 \tab 35 \tab 39 \tab 42 \tab 39 \tab 30 \tab 35 \tab 41 \tab 34 \tab 26 \tab 6 \tab 28 \tab - \tab 32 \tab 34 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 3 \tab 3 \tab 6 \tab 3 \tab 6 \tab 5 \tab 11 \tab 10 \tab 10 \tab 17 \tab 18 \tab 13 \tab 18 \tab 11 \tab 14 \tab 19 \tab 19 \tab 16 \tab 19 \tab 15 \tab 27 \tab 19 \tab 18 \tab 19 \tab 28 \tab 33 \tab 30 \tab 24 \tab 25 \tab 28 \tab 31 \tab 19 \tab 21 \tab 28 \tab 31 \tab 24 \tab 18 \tab 21 \tab 30 \tab 21 \tab 20 \tab 37 \tab 27 \tab 26 \tab 25 \tab 25 \tab 32 \tab 33 \tab 29 \tab 35 \tab 39 \tab 20 \tab 25 \tab 40 \tab 25 \tab 34 \tab 35 \tab 48 \tab 34 \tab 28 \tab 36 \tab 7 \tab 36 \tab - \tab 25 \tab 38 \tab 26 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 6 \tab 1 \tab 4 \tab 4 \tab 3 \tab 9 \tab 7 \tab 14 \tab 8 \tab 12 \tab 11 \tab 11 \tab 12 \tab 15 \tab 15 \tab 20 \tab 14 \tab 11 \tab 16 \tab 13 \tab 18 \tab 22 \tab 18 \tab 13 \tab 18 \tab 16 \tab 20 \tab 20 \tab 26 \tab 32 \tab 20 \tab 17 \tab 22 \tab 22 \tab 23 \tab 25 \tab 21 \tab 22 \tab 20 \tab 17 \tab 13 \tab 13 \tab 21 \tab 28 \tab 27 \tab 23 \tab 33 \tab 28 \tab 27 \tab 27 \tab 22 \tab 35 \tab 34 \tab 34 \tab 44 \tab 44 \tab 39 \tab 37 \tab 40 \tab 35 \tab 51 \tab 40 \tab 41 \tab 7 \tab 42 \tab - \tab 29 \tab 32 \tab 28 \tab 25 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab 4 \tab 4 \tab 2 \tab 5 \tab 6 \tab 6 \tab 6 \tab 5 \tab 5 \tab 4 \tab 13 \tab 12 \tab 9 \tab 9 \tab 9 \tab 11 \tab 20 \tab 18 \tab 18 \tab 20 \tab 26 \tab 23 \tab 26 \tab 16 \tab 24 \tab 16 \tab 19 \tab 18 \tab 22 \tab 24 \tab 27 \tab 22 \tab 20 \tab 21 \tab 20 \tab 16 \tab 20 \tab 20 \tab 28 \tab 19 \tab 19 \tab 17 \tab 20 \tab 26 \tab 15 \tab 20 \tab 21 \tab 25 \tab 18 \tab 19 \tab 35 \tab 32 \tab 28 \tab 31 \tab 46 \tab 38 \tab 34 \tab 38 \tab 38 \tab 34 \tab 42 \tab 27 \tab 9 \tab 22 \tab - \tab 39 \tab 30 \tab 38 \tab 31 \tab 26 \tab 28 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 10 \tab 3 \tab 4 \tab 5 \tab 5 \tab 6 \tab 5 \tab 8 \tab 8 \tab 10 \tab 12 \tab 9 \tab 14 \tab 12 \tab 15 \tab 15 \tab 21 \tab 13 \tab 22 \tab 21 \tab 21 \tab 29 \tab 23 \tab 19 \tab 20 \tab 22 \tab 25 \tab 19 \tab 31 \tab 24 \tab 29 \tab 31 \tab 25 \tab 38 \tab 21 \tab 31 \tab 22 \tab 30 \tab 20 \tab 24 \tab 27 \tab 24 \tab 24 \tab 26 \tab 17 \tab 28 \tab 24 \tab 25 \tab 29 \tab 30 \tab 27 \tab 33 \tab 39 \tab 32 \tab 38 \tab 37 \tab 31 \tab 38 \tab 47 \tab 51 \tab 55 \tab 51 \tab 15 \tab 61 \tab - \tab 47 \tab 57 \tab 41 \tab 45 \tab 35 \tab 40 \tab 25 \tab 33 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab 2 \tab 5 \tab 6 \tab 7 \tab 11 \tab 5 \tab 8 \tab 9 \tab 10 \tab 10 \tab 10 \tab 12 \tab 14 \tab 20 \tab 14 \tab 17 \tab 16 \tab 15 \tab 27 \tab 13 \tab 26 \tab 18 \tab 19 \tab 19 \tab 21 \tab 29 \tab 16 \tab 32 \tab 21 \tab 30 \tab 11 \tab 24 \tab 31 \tab 17 \tab 14 \tab 17 \tab 21 \tab 16 \tab 20 \tab 23 \tab 25 \tab 29 \tab 22 \tab 26 \tab 23 \tab 25 \tab 25 \tab 32 \tab 37 \tab 31 \tab 38 \tab 38 \tab 36 \tab 41 \tab 50 \tab 33 \tab 43 \tab 7 \tab 48 \tab - \tab 59 \tab 56 \tab 45 \tab 36 \tab 34 \tab 28 \tab 21 \tab 12 \tab 23 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab 2 \tab 1 \tab 2 \tab 2 \tab 4 \tab 5 \tab 8 \tab 12 \tab 6 \tab 4 \tab 6 \tab 13 \tab 14 \tab 9 \tab 15 \tab 6 \tab 14 \tab 12 \tab 18 \tab 9 \tab 24 \tab 25 \tab 14 \tab 18 \tab 13 \tab 15 \tab 20 \tab 20 \tab 27 \tab 16 \tab 24 \tab 19 \tab 19 \tab 19 \tab 17 \tab 14 \tab 21 \tab 16 \tab 8 \tab 21 \tab 22 \tab 19 \tab 18 \tab 20 \tab 23 \tab 25 \tab 21 \tab 25 \tab 28 \tab 24 \tab 31 \tab 38 \tab 41 \tab 37 \tab 31 \tab 39 \tab 37 \tab 28 \tab 28 \tab 6 \tab 33 \tab - \tab 37 \tab 44 \tab 32 \tab 39 \tab 47 \tab 50 \tab 29 \tab 37 \tab 23 \tab 20 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 7 \tab 1 \tab 7 \tab 6 \tab 7 \tab 6 \tab 12 \tab 8 \tab 8 \tab 12 \tab 11 \tab 10 \tab 13 \tab 11 \tab 12 \tab 12 \tab 16 \tab 22 \tab 21 \tab 15 \tab 23 \tab 16 \tab 24 \tab 17 \tab 27 \tab 23 \tab 20 \tab 20 \tab 27 \tab 29 \tab 20 \tab 25 \tab 13 \tab 20 \tab 20 \tab 15 \tab 34 \tab 21 \tab 23 \tab 18 \tab 16 \tab 25 \tab 18 \tab 20 \tab 22 \tab 23 \tab 25 \tab 30 \tab 30 \tab 32 \tab 36 \tab 31 \tab 32 \tab 31 \tab 33 \tab 49 \tab 30 \tab 7 \tab 37 \tab - \tab 28 \tab 40 \tab 47 \tab 37 \tab 41 \tab 24 \tab 33 \tab 26 \tab 25 \tab 30 \tab 15 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 3 \tab 4 \tab 3 \tab 7 \tab 4 \tab 7 \tab 10 \tab 8 \tab 12 \tab 15 \tab 13 \tab 11 \tab 6 \tab 16 \tab 12 \tab 15 \tab 14 \tab 22 \tab 11 \tab 20 \tab 22 \tab 25 \tab 20 \tab 15 \tab 18 \tab 27 \tab 14 \tab 21 \tab 20 \tab 10 \tab 15 \tab 19 \tab 15 \tab 20 \tab 14 \tab 12 \tab 25 \tab 22 \tab 19 \tab 16 \tab 14 \tab 25 \tab 21 \tab 25 \tab 28 \tab 28 \tab 29 \tab 27 \tab 37 \tab 32 \tab 31 \tab 32 \tab 32 \tab 39 \tab 26 \tab 45 \tab 7 \tab 37 \tab - \tab 42 \tab 35 \tab 42 \tab 39 \tab 31 \tab 43 \tab 28 \tab 22 \tab 26 \tab 22 \tab 19 \tab 18 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 8 \tab 4 \tab 3 \tab 8 \tab 6 \tab 5 \tab 7 \tab 4 \tab 7 \tab 7 \tab 8 \tab 16 \tab 14 \tab 12 \tab 15 \tab 12 \tab 22 \tab 17 \tab 23 \tab 20 \tab 20 \tab 22 \tab 18 \tab 26 \tab 28 \tab 23 \tab 22 \tab 26 \tab 15 \tab 19 \tab 20 \tab 23 \tab 20 \tab 13 \tab 26 \tab 18 \tab 30 \tab 20 \tab 26 \tab 21 \tab 19 \tab 30 \tab 19 \tab 38 \tab 43 \tab 27 \tab 32 \tab 27 \tab 28 \tab 41 \tab 41 \tab 40 \tab 37 \tab 43 \tab 46 \tab 62 \tab 35 \tab 12 \tab 41 \tab - \tab 42 \tab 53 \tab 44 \tab 44 \tab 43 \tab 37 \tab 38 \tab 34 \tab 47 \tab 32 \tab 31 \tab 24 \tab 30 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 11 \tab 2 \tab 4 \tab 3 \tab 5 \tab 4 \tab 8 \tab 3 \tab 5 \tab 3 \tab 13 \tab 18 \tab 10 \tab 16 \tab 17 \tab 15 \tab 12 \tab 15 \tab 18 \tab 18 \tab 20 \tab 23 \tab 18 \tab 19 \tab 14 \tab 22 \tab 19 \tab 20 \tab 22 \tab 22 \tab 14 \tab 19 \tab 10 \tab 17 \tab 15 \tab 10 \tab 14 \tab 12 \tab 18 \tab 15 \tab 22 \tab 11 \tab 18 \tab 16 \tab 28 \tab 33 \tab 28 \tab 34 \tab 33 \tab 31 \tab 22 \tab 39 \tab 40 \tab 28 \tab 42 \tab 28 \tab 4 \tab 30 \tab - \tab 24 \tab 35 \tab 36 \tab 30 \tab 42 \tab 38 \tab 36 \tab 40 \tab 29 \tab 31 \tab 31 \tab 28 \tab 24 \tab 24 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 9 \tab 5 \tab 3 \tab 2 \tab 5 \tab 2 \tab 4 \tab 10 \tab 9 \tab 6 \tab 8 \tab 12 \tab 14 \tab 11 \tab 17 \tab 15 \tab 8 \tab 22 \tab 8 \tab 18 \tab 18 \tab 26 \tab 19 \tab 24 \tab 24 \tab 12 \tab 20 \tab 16 \tab 18 \tab 21 \tab 13 \tab 15 \tab 14 \tab 16 \tab 24 \tab 17 \tab 17 \tab 15 \tab 31 \tab 27 \tab 24 \tab 24 \tab 28 \tab 22 \tab 28 \tab 23 \tab 25 \tab 35 \tab 29 \tab 29 \tab 28 \tab 40 \tab 29 \tab 43 \tab 32 \tab 4 \tab 29 \tab - \tab 39 \tab 36 \tab 42 \tab 34 \tab 36 \tab 47 \tab 33 \tab 23 \tab 35 \tab 28 \tab 29 \tab 23 \tab 33 \tab 21 \tab 26 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 10 \tab 4 \tab 2 \tab 2 \tab 4 \tab 6 \tab 9 \tab 5 \tab 9 \tab 6 \tab 11 \tab 12 \tab 14 \tab 9 \tab 9 \tab 21 \tab 20 \tab 16 \tab 8 \tab 14 \tab 19 \tab 17 \tab 16 \tab 18 \tab 18 \tab 17 \tab 17 \tab 18 \tab 17 \tab 15 \tab 12 \tab 13 \tab 15 \tab 21 \tab 8 \tab 15 \tab 14 \tab 15 \tab 14 \tab 14 \tab 18 \tab 19 \tab 25 \tab 20 \tab 23 \tab 25 \tab 34 \tab 26 \tab 31 \tab 33 \tab 21 \tab 40 \tab 35 \tab 35 \tab - \tab 33 \tab - \tab 36 \tab 35 \tab 36 \tab 32 \tab 35 \tab 34 \tab 33 \tab 24 \tab 16 \tab 27 \tab 26 \tab 19 \tab 37 \tab 19 \tab 22 \tab 15 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 8 \tab 2 \tab 4 \tab - \tab 6 \tab 5 \tab 5 \tab 13 \tab 9 \tab 11 \tab 7 \tab 11 \tab 12 \tab 13 \tab 10 \tab 17 \tab 24 \tab 19 \tab 20 \tab 15 \tab 17 \tab 18 \tab 18 \tab 13 \tab 18 \tab 15 \tab 19 \tab 17 \tab 20 \tab 16 \tab 15 \tab 18 \tab 17 \tab 12 \tab 16 \tab 18 \tab 14 \tab 20 \tab 15 \tab 20 \tab 22 \tab 29 \tab 28 \tab 12 \tab 22 \tab 30 \tab 33 \tab 30 \tab 38 \tab 36 \tab 35 \tab 37 \tab 33 \tab 3 \tab 35 \tab - \tab 55 \tab 35 \tab 38 \tab 44 \tab 35 \tab 35 \tab 27 \tab 33 \tab 29 \tab 33 \tab 28 \tab 24 \tab 28 \tab 35 \tab 38 \tab 18 \tab 9 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab 3 \tab 6 \tab 5 \tab 4 \tab 4 \tab 3 \tab 12 \tab 9 \tab 6 \tab 7 \tab 14 \tab 15 \tab 14 \tab 18 \tab 19 \tab 20 \tab 20 \tab 17 \tab 12 \tab 20 \tab 18 \tab 11 \tab 20 \tab 18 \tab 14 \tab 19 \tab 18 \tab 12 \tab 17 \tab 16 \tab 25 \tab 19 \tab 13 \tab 23 \tab 19 \tab 18 \tab 22 \tab 27 \tab 24 \tab 29 \tab 20 \tab 36 \tab 33 \tab 33 \tab 29 \tab 40 \tab 32 \tab 39 \tab 40 \tab 41 \tab 5 \tab 44 \tab - \tab 38 \tab 45 \tab 44 \tab 28 \tab 36 \tab 30 \tab 36 \tab 39 \tab 33 \tab 46 \tab 23 \tab 27 \tab 35 \tab 26 \tab 28 \tab 16 \tab 24 \tab 20 \tab 18 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 20 \tab 4 \tab 6 \tab 7 \tab 12 \tab 9 \tab 9 \tab 16 \tab 11 \tab 21 \tab 20 \tab 21 \tab 20 \tab 32 \tab 28 \tab 31 \tab 26 \tab 25 \tab 32 \tab 28 \tab 30 \tab 32 \tab 30 \tab 28 \tab 27 \tab 23 \tab 42 \tab 25 \tab 26 \tab 32 \tab 43 \tab 29 \tab 40 \tab 38 \tab 32 \tab 44 \tab 45 \tab 56 \tab 34 \tab 55 \tab 66 \tab 55 \tab 71 \tab 60 \tab 48 \tab 65 \tab 69 \tab 59 \tab 71 \tab 59 \tab 6 \tab 65 \tab - \tab 80 \tab 60 \tab 80 \tab 76 \tab 88 \tab 78 \tab 83 \tab 74 \tab 69 \tab 56 \tab 68 \tab 57 \tab 55 \tab 62 \tab 51 \tab 42 \tab 31 \tab 29 \tab 34 \tab 27 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 18 \tab 7 \tab 7 \tab 11 \tab 5 \tab 9 \tab 11 \tab 11 \tab 8 \tab 16 \tab 15 \tab 17 \tab 24 \tab 16 \tab 18 \tab 27 \tab 17 \tab 27 \tab 13 \tab 39 \tab 24 \tab 31 \tab 21 \tab 25 \tab 27 \tab 24 \tab 29 \tab 38 \tab 26 \tab 33 \tab 33 \tab 27 \tab 25 \tab 41 \tab 31 \tab 45 \tab 55 \tab 40 \tab 49 \tab 48 \tab 69 \tab 66 \tab 57 \tab 52 \tab 54 \tab 53 \tab 74 \tab 67 \tab 6 \tab 63 \tab - \tab 68 \tab 58 \tab 82 \tab 89 \tab 61 \tab 61 \tab 64 \tab 74 \tab 75 \tab 60 \tab 72 \tab 59 \tab 68 \tab 53 \tab 58 \tab 66 \tab 66 \tab 58 \tab 45 \tab 37 \tab 40 \tab 34 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 12 \tab 6 \tab 12 \tab 9 \tab 13 \tab 7 \tab 14 \tab 12 \tab 12 \tab 15 \tab 18 \tab 12 \tab 21 \tab 20 \tab 27 \tab 20 \tab 41 \tab 26 \tab 24 \tab 30 \tab 24 \tab 24 \tab 29 \tab 23 \tab 27 \tab 28 \tab 22 \tab 33 \tab 26 \tab 31 \tab 26 \tab 34 \tab 33 \tab 42 \tab 45 \tab 37 \tab 34 \tab 40 \tab 72 \tab 42 \tab 50 \tab 51 \tab 49 \tab 52 \tab 54 \tab 67 \tab 5 \tab 61 \tab - \tab 63 \tab 63 \tab 70 \tab 71 \tab 63 \tab 71 \tab 71 \tab 79 \tab 83 \tab 74 \tab 66 \tab 56 \tab 53 \tab 57 \tab 77 \tab 50 \tab 61 \tab 50 \tab 37 \tab 42 \tab 37 \tab 48 \tab 36 \tab 36 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 17 \tab 5 \tab 5 \tab 11 \tab 15 \tab 17 \tab 17 \tab 15 \tab 17 \tab 17 \tab 22 \tab 22 \tab 20 \tab 20 \tab 19 \tab 20 \tab 28 \tab 28 \tab 21 \tab 22 \tab 20 \tab 28 \tab 27 \tab 32 \tab 18 \tab 20 \tab 29 \tab 29 \tab 38 \tab 32 \tab 34 \tab 29 \tab 29 \tab 42 \tab 26 \tab 46 \tab 51 \tab 58 \tab 42 \tab 51 \tab 48 \tab 62 \tab 54 \tab 55 \tab 8 \tab 69 \tab - \tab 70 \tab 70 \tab 73 \tab 60 \tab 73 \tab 53 \tab 76 \tab 63 \tab 62 \tab 55 \tab 55 \tab 45 \tab 77 \tab 62 \tab 62 \tab 56 \tab 45 \tab 47 \tab 58 \tab 53 \tab 52 \tab 48 \tab 39 \tab 37 \tab 40 \tab 44 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 13 \tab 6 \tab 10 \tab 6 \tab 11 \tab 14 \tab 15 \tab 11 \tab 13 \tab 16 \tab 9 \tab 15 \tab 19 \tab 31 \tab 26 \tab 24 \tab 22 \tab 23 \tab 36 \tab 34 \tab 27 \tab 26 \tab 30 \tab 20 \tab 30 \tab 38 \tab 34 \tab 32 \tab 31 \tab 30 \tab 33 \tab 33 \tab 35 \tab 47 \tab 39 \tab 45 \tab 45 \tab 58 \tab 52 \tab 47 \tab 43 \tab 61 \tab 14 \tab 52 \tab - \tab 70 \tab 51 \tab 47 \tab 63 \tab 44 \tab 54 \tab 60 \tab 54 \tab 49 \tab 64 \tab 55 \tab 59 \tab 57 \tab 76 \tab 62 \tab 53 \tab 61 \tab 43 \tab 51 \tab 48 \tab 54 \tab 61 \tab 58 \tab 68 \tab 46 \tab 40 \tab 32 \tab 25 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 14 \tab 7 \tab 4 \tab 8 \tab 11 \tab 7 \tab 11 \tab 15 \tab 15 \tab 9 \tab 22 \tab 17 \tab 8 \tab 14 \tab 25 \tab 23 \tab 12 \tab 18 \tab 22 \tab 22 \tab 22 \tab 22 \tab 41 \tab 29 \tab 32 \tab 26 \tab 36 \tab 43 \tab 43 \tab 44 \tab 39 \tab 36 \tab 53 \tab 56 \tab 50 \tab 36 \tab 40 \tab 58 \tab 53 \tab 59 \tab 9 \tab 61 \tab - \tab 57 \tab 62 \tab 58 \tab 57 \tab 68 \tab 60 \tab 49 \tab 58 \tab 73 \tab 51 \tab 54 \tab 45 \tab 58 \tab 62 \tab 70 \tab 58 \tab 52 \tab 52 \tab 55 \tab 43 \tab 69 \tab 52 \tab 59 \tab 55 \tab 50 \tab 47 \tab 36 \tab 38 \tab 32 \tab 33 \tab 27 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 29 \tab 11 \tab 7 \tab 17 \tab 18 \tab 17 \tab 19 \tab 21 \tab 25 \tab 27 \tab 22 \tab 28 \tab 33 \tab 31 \tab 39 \tab 33 \tab 29 \tab 48 \tab 39 \tab 41 \tab 42 \tab 41 \tab 45 \tab 53 \tab 42 \tab 56 \tab 62 \tab 58 \tab 57 \tab 75 \tab 78 \tab 81 \tab 58 \tab 68 \tab 86 \tab 96 \tab 89 \tab 78 \tab 18 \tab 76 \tab - \tab 93 \tab 97 \tab 109 \tab 110 \tab 90 \tab 79 \tab 100 \tab 95 \tab 99 \tab 94 \tab 95 \tab 95 \tab 94 \tab 89 \tab 105 \tab 80 \tab 91 \tab 88 \tab 97 \tab 67 \tab 84 \tab 79 \tab 80 \tab 86 \tab 83 \tab 74 \tab 61 \tab 63 \tab 72 \tab 52 \tab 50 \tab 47 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 21 \tab 1 \tab 6 \tab 6 \tab 3 \tab 7 \tab 8 \tab 12 \tab 13 \tab 15 \tab 12 \tab 17 \tab 11 \tab 17 \tab 18 \tab 23 \tab 11 \tab 19 \tab 16 \tab 21 \tab 21 \tab 16 \tab 27 \tab 23 \tab 31 \tab 25 \tab 19 \tab 34 \tab 36 \tab 40 \tab 20 \tab 24 \tab 41 \tab 42 \tab 33 \tab 38 \tab 10 \tab 42 \tab - \tab 37 \tab 43 \tab 51 \tab 41 \tab 38 \tab 49 \tab 41 \tab 34 \tab 40 \tab 35 \tab 28 \tab 36 \tab 39 \tab 33 \tab 38 \tab 44 \tab 45 \tab 41 \tab 40 \tab 36 \tab 22 \tab 47 \tab 27 \tab 31 \tab 40 \tab 37 \tab 41 \tab 26 \tab 34 \tab 37 \tab 27 \tab 21 \tab 28 \tab 31 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 20 \tab 9 \tab 4 \tab 7 \tab 6 \tab 12 \tab 12 \tab 13 \tab 10 \tab 17 \tab 11 \tab 8 \tab 14 \tab 17 \tab 15 \tab 15 \tab 8 \tab 18 \tab 15 \tab 24 \tab 26 \tab 27 \tab 37 \tab 20 \tab 28 \tab 19 \tab 35 \tab 39 \tab 27 \tab 44 \tab 41 \tab 38 \tab 31 \tab 40 \tab 3 \tab 32 \tab - \tab 40 \tab 37 \tab 37 \tab 39 \tab 43 \tab 45 \tab 27 \tab 29 \tab 40 \tab 38 \tab 39 \tab 42 \tab 34 \tab 38 \tab 36 \tab 35 \tab 31 \tab 39 \tab 30 \tab 45 \tab 32 \tab 36 \tab 33 \tab 42 \tab 40 \tab 45 \tab 30 \tab 42 \tab 29 \tab 48 \tab 31 \tab 38 \tab 19 \tab 35 \tab 24 \tab 24 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 15 \tab 7 \tab 10 \tab 10 \tab 5 \tab 13 \tab 6 \tab 10 \tab 12 \tab 10 \tab 11 \tab 15 \tab 15 \tab 17 \tab 16 \tab 17 \tab 20 \tab 15 \tab 29 \tab 17 \tab 20 \tab 24 \tab 23 \tab 19 \tab 38 \tab 31 \tab 35 \tab 31 \tab 37 \tab 27 \tab 31 \tab 25 \tab 5 \tab 31 \tab - \tab 29 \tab 47 \tab 33 \tab 32 \tab 37 \tab 40 \tab 32 \tab 43 \tab 38 \tab 36 \tab 34 \tab 35 \tab 29 \tab 29 \tab 43 \tab 40 \tab 30 \tab 42 \tab 37 \tab 40 \tab 33 \tab 35 \tab 34 \tab 39 \tab 47 \tab 35 \tab 47 \tab 28 \tab 41 \tab 32 \tab 29 \tab 31 \tab 27 \tab 30 \tab 31 \tab 31 \tab 19 \tab 20 \tab 12 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 19 \tab 11 \tab 8 \tab 9 \tab 6 \tab 13 \tab 11 \tab 8 \tab 14 \tab 16 \tab 26 \tab 28 \tab 17 \tab 24 \tab 21 \tab 25 \tab 22 \tab 20 \tab 28 \tab 32 \tab 23 \tab 21 \tab 41 \tab 35 \tab 40 \tab 38 \tab 42 \tab 44 \tab 38 \tab 46 \tab 9 \tab 47 \tab - \tab 60 \tab 43 \tab 60 \tab 52 \tab 53 \tab 60 \tab 37 \tab 45 \tab 54 \tab 39 \tab 32 \tab 43 \tab 34 \tab 41 \tab 39 \tab 54 \tab 35 \tab 54 \tab 46 \tab 52 \tab 47 \tab 54 \tab 40 \tab 40 \tab 46 \tab 47 \tab 55 \tab 49 \tab 57 \tab 51 \tab 31 \tab 41 \tab 47 \tab 48 \tab 31 \tab 30 \tab 28 \tab 24 \tab 26 \tab 25 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 22 \tab 3 \tab 9 \tab 12 \tab 11 \tab 11 \tab 22 \tab 11 \tab 21 \tab 25 \tab 21 \tab 26 \tab 23 \tab 25 \tab 19 \tab 24 \tab 22 \tab 26 \tab 27 \tab 32 \tab 42 \tab 43 \tab 49 \tab 47 \tab 44 \tab 34 \tab 43 \tab 56 \tab 9 \tab 56 \tab - \tab 53 \tab 58 \tab 67 \tab 70 \tab 48 \tab 48 \tab 53 \tab 60 \tab 65 \tab 49 \tab 48 \tab 41 \tab 46 \tab 55 \tab 45 \tab 27 \tab 52 \tab 45 \tab 50 \tab 36 \tab 46 \tab 42 \tab 44 \tab 58 \tab 52 \tab 56 \tab 49 \tab 54 \tab 42 \tab 57 \tab 54 \tab 56 \tab 42 \tab 58 \tab 42 \tab 56 \tab 35 \tab 49 \tab 44 \tab 33 \tab 26 \tab 34 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 29 \tab 8 \tab 5 \tab 14 \tab 12 \tab 10 \tab 7 \tab 12 \tab 9 \tab 12 \tab 9 \tab 11 \tab 28 \tab 19 \tab 23 \tab 29 \tab 20 \tab 22 \tab 33 \tab 45 \tab 41 \tab 39 \tab 27 \tab 36 \tab 39 \tab 33 \tab 7 \tab 37 \tab - \tab 24 \tab 39 \tab 46 \tab 46 \tab 37 \tab 39 \tab 50 \tab 48 \tab 35 \tab 48 \tab 29 \tab 41 \tab 37 \tab 31 \tab 33 \tab 30 \tab 42 \tab 29 \tab 40 \tab 39 \tab 40 \tab 40 \tab 41 \tab 55 \tab 38 \tab 43 \tab 40 \tab 40 \tab 65 \tab 42 \tab 43 \tab 34 \tab 50 \tab 43 \tab 45 \tab 39 \tab 31 \tab 43 \tab 38 \tab 29 \tab 40 \tab 27 \tab 15 \tab 26 \tab 22 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 26 \tab 5 \tab 5 \tab 14 \tab 16 \tab 21 \tab 17 \tab 18 \tab 31 \tab 30 \tab 40 \tab 32 \tab 37 \tab 34 \tab 57 \tab 69 \tab 48 \tab 52 \tab 55 \tab 71 \tab 73 \tab 72 \tab 89 \tab 333 \tab 58 \tab - \tab 68 \tab 68 \tab 77 \tab 77 \tab 76 \tab 70 \tab 81 \tab 60 \tab 79 \tab 63 \tab 65 \tab 53 \tab 65 \tab 53 \tab 69 \tab 51 \tab 54 \tab 47 \tab 54 \tab 61 \tab 52 \tab 60 \tab 56 \tab 77 \tab 76 \tab 63 \tab 65 \tab 59 \tab 74 \tab 61 \tab 62 \tab 69 \tab 82 \tab 59 \tab 56 \tab 61 \tab 63 \tab 59 \tab 35 \tab 44 \tab 46 \tab 38 \tab 30 \tab 31 \tab 24 \tab 18 \tab 14 \tab 4 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 19 \tab 4 \tab 5 \tab 11 \tab 8 \tab 21 \tab 13 \tab 15 \tab 15 \tab 22 \tab 30 \tab 26 \tab 32 \tab 31 \tab 34 \tab 43 \tab 34 \tab 54 \tab 40 \tab 67 \tab 56 \tab 68 \tab 208 \tab 44 \tab - \tab 67 \tab 65 \tab 50 \tab 48 \tab 60 \tab 57 \tab 75 \tab 51 \tab 44 \tab 53 \tab 56 \tab 60 \tab 50 \tab 38 \tab 47 \tab 55 \tab 40 \tab 51 \tab 46 \tab 53 \tab 52 \tab 56 \tab 55 \tab 45 \tab 67 \tab 48 \tab 62 \tab 72 \tab 57 \tab 47 \tab 69 \tab 58 \tab 61 \tab 69 \tab 76 \tab 68 \tab 60 \tab 61 \tab 57 \tab 46 \tab 58 \tab 45 \tab 45 \tab 53 \tab 46 \tab 46 \tab 48 \tab 29 \tab 22 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 25 \tab 4 \tab 4 \tab 7 \tab 13 \tab 11 \tab 14 \tab 16 \tab 14 \tab 18 \tab 31 \tab 15 \tab 30 \tab 36 \tab 34 \tab 43 \tab 52 \tab 39 \tab 53 \tab 41 \tab 61 \tab 133 \tab 59 \tab - \tab 60 \tab 65 \tab 60 \tab 71 \tab 58 \tab 60 \tab 48 \tab 63 \tab 56 \tab 45 \tab 35 \tab 57 \tab 45 \tab 67 \tab 48 \tab 43 \tab 40 \tab 35 \tab 46 \tab 53 \tab 56 \tab 51 \tab 43 \tab 55 \tab 54 \tab 51 \tab 66 \tab 54 \tab 51 \tab 67 \tab 56 \tab 67 \tab 64 \tab 52 \tab 62 \tab 51 \tab 68 \tab 55 \tab 41 \tab 39 \tab 41 \tab 50 \tab 47 \tab 49 \tab 39 \tab 32 \tab 36 \tab 24 \tab 32 \tab 32 \tab 16 \tab 3309 \cr
   Total \tab 2 \tab 3 \tab 7 \tab 13 \tab 10 \tab 21 \tab 27 \tab 27 \tab 31 \tab 34 \tab 61 \tab 55 \tab 73 \tab 72 \tab 105 \tab 123 \tab 130 \tab 138 \tab 163 \tab 173 \tab 182 \tab 219 \tab 207 \tab 263 \tab 246 \tab 330 \tab 285 \tab 318 \tab 348 \tab 371 \tab 417 \tab 428 \tab 455 \tab 483 \tab 500 \tab 530 \tab 520 \tab 582 \tab 572 \tab 605 \tab 618 \tab 607 \tab 650 \tab 660 \tab 614 \tab 646 \tab 651 \tab 612 \tab 656 \tab 640 \tab 751 \tab 747 \tab 773 \tab 777 \tab 825 \tab 838 \tab 864 \tab 921 \tab 1034 \tab 1106 \tab 1131 \tab 1047 \tab 1232 \tab 1500 \tab 1429 \tab 1322 \tab 1443 \tab 1509 \tab 1558 \tab 1533 \tab 1518 \tab 902 \tab 1469 \tab 4 \tab 1559 \tab 1541 \tab 1547 \tab 1469 \tab 1385 \tab 1358 \tab 1288 \tab 1253 \tab 1253 \tab 1139 \tab 1035 \tab 991 \tab 1029 \tab 976 \tab 1000 \tab 838 \tab 813 \tab 773 \tab 784 \tab 737 \tab 725 \tab 743 \tab 652 \tab 724 \tab 685 \tab 630 \tab 591 \tab 550 \tab 558 \tab 527 \tab 479 \tab 462 \tab 439 \tab 425 \tab 373 \tab 360 \tab 314 \tab 311 \tab 253 \tab 216 \tab 217 \tab 194 \tab 144 \tab 159 \tab 131 \tab 96 \tab 98 \tab 57 \tab 54 \tab 32 \tab 16 \tab 75699 \cr
}
}

\seealso{
Other Cohort and Birth: 
\code{\link{marcohrt}},
\code{\link{zodiac}}
}
\concept{Age}
\concept{Cohort and Birth}
\keyword{variable}
