% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{cohrs2}
\alias{cohrs2}
\title{No. of hours partner usually works a week}
\source{
General Social Survey https://gss.norc.org
}
\description{
cohrs2
}
\details{
Question Last week was your (partner) working full time, part time, going to school, keeping house, or what?
B. IF WITH A JOB, BUT NOT AT WORK: How many hours a week does (he/she) usually work, at all jobs?
}
\section{Values}{

\itemize{
\item \code{89} 89+ hours
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6156/vshow}{GSS Data Explorer page for \code{cohrs2}}.
}

\seealso{
Other Core: 
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{sppres}},
\code{\link{sppres80}},
\code{\link{spwrkslf}}
}
\concept{Core}
\keyword{variable}
