% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{cojail}
\alias{cojail}
\title{Jail draft resister's?}
\source{
General Social Survey https://gss.norc.org
}
\description{
cojail
}
\details{
Question Even though they are no longer drafted for military service, young men are still required by law to register for the draft when they become 18 years old. If a young man refuses to register for the draft, do you think he should be punished in any way?
A. If a young man refuses to register for the draft, would you approve or disapprove of sending him to jail?
}
\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/817/vshow}{GSS Data Explorer page for \code{cojail}}.
}

\seealso{
Other Core: 
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}
}
\concept{Core}
\keyword{variable}
