% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{cojew}
\alias{cojew}
\title{What sort of jew is r's partner currently}
\source{
General Social Survey https://gss.norc.org
}
\description{
cojew
}
\details{
Question Does (s/he) consider (her/himself) Orthodox, Conservative, Reform or none of these?
}
\section{Values}{

\itemize{
\item \verb{[1]} orthodox
\item \verb{[2]} conservative
\item \verb{[3]} reform
\item \verb{[4]} none of these
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6172/vshow}{GSS Data Explorer page for \code{cojew}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab conservative \tab reform \tab no answer \tab none of these \tab skipped on web \tab 5 \tab don't know \tab Total \cr
   2012 \tab 1972 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2535 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2863 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2346 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4020 \tab - \tab 5 \tab - \tab 6 \tab 1 \tab - \tab - \tab 4032 \cr
   2022 \tab 3536 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 3544 \cr
   Total \tab 17272 \tab 2 \tab 9 \tab 3 \tab 13 \tab 1 \tab 1 \tab 2 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayerx}},
\code{\link{prayery}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
