% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{colaff}
\alias{colaff}
\title{Chance blacks get to college ahead of whites}
\source{
General Social Survey https://gss.norc.org
}
\description{
colaff
}
\details{
Question A. What do you think the chances are these days that a white person won't get admitted to a college or university program while an equally or less qualified black person gets admitted instead?  Would you say: very likely, somewhat likely, or not likely.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1622/vshow}{GSS Data Explorer page for \code{colaff}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not likely \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2235 \tab 195 \tab 311 \tab 207 \tab 38 \tab 6 \tab 2992 \cr
   Total \tab 2235 \tab 195 \tab 311 \tab 207 \tab 38 \tab 6 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{ethhist}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
