% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{colaffy}
\alias{colaffy}
\title{Chance whites excluded from college over blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
colaffy
}
\details{
Question A. What do you think the chances are these days that a white person won't get admitted to a college or university program while a less qualified black person gets admitted instead? Would you say: very likely, somewhat likely, or not likely.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1623/vshow}{GSS Data Explorer page for \code{colaffy}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{ethhist}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}

Other Affirmative Action: 
\code{\link{colaff}},
\code{\link{discaffy}},
\code{\link{jobaff}}

Other Blacks: 
\code{\link{colaff}},
\code{\link{discaffy}},
\code{\link{jobaff}}

Other Education: 
\code{\link{colaff}},
\code{\link{ethhist}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Affirmative Action}
\concept{Blacks}
\concept{Education}
\concept{Multiculturalism}
\keyword{variable}
