% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{colcomy}
\alias{colcomy}
\title{Should communist teacher be fired y}
\source{
General Social Survey https://gss.norc.org
}
\description{
colcomy
}
\details{
Question (Now, we would like to ask you some questions about a self-identified Communist...) Suppose this person is teaching in a college. Should this person be fired, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, fired
\item \verb{[2]} not fired
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7439/vshow}{GSS Data Explorer page for \code{colcomy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, fired \tab not fired \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 2675 \tab 444 \tab 885 \tab 4 \tab 24 \tab - \tab 4032 \cr
   2022 \tab 2367 \tab 407 \tab 706 \tab 54 \tab 3 \tab 7 \tab 3544 \cr
   Total \tab 5042 \tab 851 \tab 1591 \tab 58 \tab 27 \tab 7 \tab 7576 \cr
}
}

\keyword{variable}
