% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{coldeg1}
\alias{coldeg1}
\title{The highest degree r have earned}
\source{
General Social Survey https://gss.norc.org
}
\description{
coldeg1
}
\details{
Question What is the highest degree you have earned?
}
\section{Values}{

\itemize{
\item \verb{[1]} associate's
\item \verb{[2]} bachelor's
\item \verb{[3]} master's
\item \verb{[4]} mba
\item \verb{[5]} law
\item \verb{[6]} phd
\item \verb{[7]} md
\item \verb{[8]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3347/vshow}{GSS Data Explorer page for \code{coldeg1}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab associate's \tab bachelor's \tab master's \tab mba \tab law \tab phd \tab md \tab other \tab no answer \tab don't know \tab Total \cr
   2006 \tab 3747 \tab 240 \tab 331 \tab 133 \tab 5 \tab 15 \tab 24 \tab 6 \tab 8 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1493 \tab 128 \tab 254 \tab 104 \tab 13 \tab 8 \tab 15 \tab 3 \tab 4 \tab - \tab 1 \tab 2023 \cr
   2010 \tab 1772 \tab 50 \tab 125 \tab 66 \tab 4 \tab 5 \tab 14 \tab 1 \tab 4 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 1612 \tab 75 \tab 186 \tab 73 \tab 7 \tab 5 \tab 10 \tab 3 \tab 3 \tab - \tab - \tab 1974 \cr
   2014 \tab 2098 \tab 82 \tab 220 \tab 89 \tab 7 \tab 8 \tab 19 \tab 8 \tab 5 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2336 \tab 110 \tab 254 \tab 124 \tab 8 \tab 9 \tab 15 \tab 3 \tab 5 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1913 \tab 82 \tab 215 \tab 95 \tab 9 \tab 4 \tab 15 \tab 5 \tab 7 \tab 2 \tab 1 \tab 2348 \cr
   Total \tab 14971 \tab 767 \tab 1585 \tab 684 \tab 53 \tab 54 \tab 112 \tab 29 \tab 36 \tab 11 \tab 2 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
