% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{colmajr2}
\alias{colmajr2}
\title{R 2nd major in college}
\source{
General Social Survey https://gss.norc.org
}
\description{
colmajr2
}
\details{
Question A. Could you tell me your major field of study (while you were) in college?
}
\section{Values}{

\itemize{
\item \verb{[0]} none
\item \verb{[1]} english, literature
\item \verb{[2]} foreign language
\item \verb{[3]} fine arts
\item \verb{[4]} other humanities
\item \verb{[5]} economics
\item \verb{[6]} psychology
\item \verb{[7]} government-civics
\item \verb{[8]} social science-other social study
\item \verb{[9]} mathematics
\item \verb{[10]} science
\item \verb{[11]} other
\item \verb{[12]} business
\item \verb{[13]} education
\item \verb{[14]} engineering
\item \verb{[15]} other vocational
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1451/vshow}{GSS Data Explorer page for \code{colmajr2}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab foreign language \tab fine arts \tab other humanities \tab economics \tab psychology \tab social science-other social study \tab mathematics \tab science \tab other \tab business \tab education \tab engineering \tab other vocational \tab no answer \tab Total \cr
   1993 \tab 1507 \tab 2 \tab 2 \tab 1 \tab 1 \tab 5 \tab 5 \tab 6 \tab 5 \tab 16 \tab 15 \tab 17 \tab 6 \tab 10 \tab 8 \tab 1606 \cr
   Total \tab 1507 \tab 2 \tab 2 \tab 1 \tab 1 \tab 5 \tab 5 \tab 6 \tab 5 \tab 16 \tab 15 \tab 17 \tab 6 \tab 10 \tab 8 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{colmajr1}},
\code{\link{hsclass1}},
\code{\link{hsclass2}},
\code{\link{mostimp4}}
}
\concept{Culture}
\keyword{variable}
