% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{colmslm}
\alias{colmslm}
\title{Allow anti-american muslim clergymen teaching in college}
\source{
General Social Survey https://gss.norc.org
}
\description{
colmslm
}
\details{
Question Now consider a Muslim clergyman who preaches hatred of the United States.
B. Should such a person be allowed to teach in a college or university, or not?
}
\section{Values}{

\itemize{
\item \code{4} yes, allowed to teach
\item \code{5} not allowed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
   2012 \tab A/C \tab partial \cr
   2014 \tab A/C \tab partial \cr
   2016 \tab A/C \tab partial \cr
   2018 \tab A/C \tab partial \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/265/vshow}{GSS Data Explorer page for \code{colmslm}}.
}

\seealso{
Other Core: 
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{colhomo}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}

Other Civil Liberties: 
\code{\link{colhomo}},
\code{\link{libhomo}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}

Other Religion: 
\code{\link{libmslm}},
\code{\link{spkmslm}}
}
\concept{Civil Liberties}
\concept{Core}
\concept{Religion}
\keyword{variable}
