% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{colop}
\alias{colop}
\title{Should chances for college be changed?}
\source{
General Social Survey https://gss.norc.org
}
\description{
colop
}
\details{
Question How do you feel about opportunities for young people to go to college. Should opportunities be...
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} kept the same
\item \code{4} reduced a little
\item \code{5} reduced a lot
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4024/vshow}{GSS Data Explorer page for \code{colop}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}

Other Education: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}
}
\concept{Education}
\concept{ISSP Role of Gov't}
\concept{Opportunity}
\concept{Split Ballots}
\keyword{variable}
