% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{colop}
\alias{colop}
\title{Should chances for college be changed?}
\source{
General Social Survey https://gss.norc.org
}
\description{
colop
}
\details{
Question How do you feel about opportunities for young people to go to college. Should opportunities be...
}
\section{Values}{

\itemize{
\item \verb{[1]} increased a lot
\item \verb{[2]} increased a little
\item \verb{[3]} kept the same
\item \verb{[4]} reduced a little
\item \verb{[5]} reduced a lot
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4024/vshow}{GSS Data Explorer page for \code{colop}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab kept the same \tab reduced a little \tab reduced a lot \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 288 \tab 173 \tab 173 \tab 19 \tab 4 \tab 15 \tab 5 \tab 1534 \cr
   Total \tab 857 \tab 288 \tab 173 \tab 173 \tab 19 \tab 4 \tab 15 \tab 5 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hsjudge}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
