% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{color}
\alias{color}
\title{R's skin color}
\source{
General Social Survey https://gss.norc.org
}
\description{
color
}
\details{
Question Respondent''s skin color is . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} very dark brown
\item \verb{[2]} dark brown
\item \verb{[3]} medium brown
\item \verb{[4]} light brown
\item \verb{[5]} very light brown
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/403/vshow}{GSS Data Explorer page for \code{color}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very dark brown \tab dark brown \tab medium brown \tab light brown \tab very light brown \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 50 \tab 125 \tab 233 \tab 73 \tab 14 \tab 15 \tab 1860 \cr
   Total \tab 1350 \tab 50 \tab 125 \tab 233 \tab 73 \tab 14 \tab 15 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\keyword{variable}
