% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_371.R
\name{colsci}
\alias{colsci}
\title{R has taken any college-level sci course}
\source{
General Social Survey https://gss.norc.org
}
\description{
colsci
}
\details{
Question Have you ever taken any college-level science courses?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3349/vshow}{GSS Data Explorer page for \code{colsci}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 824 \tab 1035 \tab 4 \tab 1 \tab 4510 \cr
   2008 \tab 518 \tab 568 \tab 929 \tab 4 \tab 4 \tab 2023 \cr
   2010 \tab 1328 \tab 334 \tab 379 \tab - \tab 3 \tab 2044 \cr
   2012 \tab 972 \tab 414 \tab 578 \tab 6 \tab 4 \tab 1974 \cr
   2014 \tab 1299 \tab 500 \tab 736 \tab 2 \tab 1 \tab 2538 \cr
   2016 \tab 1477 \tab 572 \tab 811 \tab 1 \tab 6 \tab 2867 \cr
   2018 \tab 1173 \tab 513 \tab 652 \tab 6 \tab 4 \tab 2348 \cr
   Total \tab 9413 \tab 3725 \tab 5120 \tab 23 \tab 23 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{colscinm}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsmath}},
\code{\link{hsphys}},
\code{\link{sciagrgw}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
