% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{colsoc}
\alias{colsoc}
\title{Allow socialist to teach}
\source{
General Social Survey https://gss.norc.org
}
\description{
colsoc
}
\details{
Question Or consider a person who favored government ownership of all the railroads and all big industries....
B. Should such a person be allowed to teach in a college or university, or not?
}
\section{Values}{

\itemize{
\item \verb{[4]} allowed
\item \verb{[5]} not allowed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973, 1974 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/250/vshow}{GSS Data Explorer page for \code{colsoc}}.

Counts by year:\tabular{lllllll}{
   year \tab allowed \tab not allowed \tab don't know \tab no answer \tab iap \tab Total \cr
   1972 \tab 899 \tab 595 \tab 114 \tab 5 \tab - \tab 1613 \cr
   1973 \tab 863 \tab 558 \tab 78 \tab 5 \tab - \tab 1504 \cr
   1974 \tab 847 \tab 552 \tab 81 \tab 4 \tab - \tab 1484 \cr
   Total \tab 2609 \tab 1705 \tab 273 \tab 14 \tab 0 \tab 4601 \cr
}
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}},
\code{\link{tax}}
}
\concept{Core}
\keyword{variable}
