% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{colsoc}
\alias{colsoc}
\title{Allow socialist to teach}
\source{
General Social Survey https://gss.norc.org
}
\description{
colsoc
}
\details{
Question Or consider a person who favored government ownership of all the railroads and all big industries....
B. Should such a person be allowed to teach in a college or university, or not?
}
\section{Values}{

\itemize{
\item \code{4} allowed
\item \code{5} not allowed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
   1973 \tab - \tab full \cr
   1974 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/250/vshow}{GSS Data Explorer page for \code{colsoc}}.
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}},
\code{\link{tax}}

Other Civil Liberties: 
\code{\link{colath}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}}
}
\concept{Civil Liberties}
\concept{Core}
\keyword{variable}
