% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_571.R
\name{comemail}
\alias{comemail}
\title{Punishment emailing documents unauthorized per}
\source{
General Social Survey https://gss.norc.org
}
\description{
comemail
}
\details{
Question Now I’m going to ask you about several offenses involving the use of a government computer system. For each case, indicate what you would consider to be the appropriate punishment when the offender is a military service member or government employee.
E. E-mailing SECRET or TOP SECRET government documents or data files to an unauthorized person.
}
\section{Values}{

\itemize{
\item \code{1} reprimand
\item \code{2} dismissal from the military/fired by the government
\item \code{3} one year in prison
\item \code{4} 5-9 years in prison
\item \code{5} 10-20 years in prison
\item \code{6} life in prison
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5241/vshow}{GSS Data Explorer page for \code{comemail}}.
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askcrime}},
\code{\link{askdrink}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{askmentl}},
\code{\link{asksexor}},
\code{\link{bugging}},
\code{\link{camwrk}},
\code{\link{chkfinan}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{chktravl}},
\code{\link{comdata}},
\code{\link{compfin}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{emailhme}},
\code{\link{emailwrk}},
\code{\link{finanqs}},
\code{\link{forspy}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{knomentl}},
\code{\link{leakinfo}},
\code{\link{lietest}},
\code{\link{nuclrwar}},
\code{\link{punarms}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}},
\code{\link{rptcowrk}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secdocs}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{sectech}},
\code{\link{secterr}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{srchwrk}},
\code{\link{takearms}},
\code{\link{taketrck}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{testdrug}},
\code{\link{usspy}},
\code{\link{usterror}}
}
\concept{National Security}
\keyword{variable}
