% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{comfort}
\alias{comfort}
\title{Religion helps people to gain comfort in times of trouble and sorrow}
\source{
General Social Survey https://gss.norc.org
}
\description{
comfort
}
\details{
Question Do you agree or disagree that…
C. ...that practicing a religion helps people to gain comfort in times of trouble and sorrow?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4714/vshow}{GSS Data Explorer page for \code{comfort}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 626 \tab 663 \tab 51 \tab 5 \tab 6 \tab 9 \tab 5 \tab 2023 \cr
   2018 \tab 1173 \tab 484 \tab 602 \tab 55 \tab 17 \tab 4 \tab 2 \tab 11 \tab 2348 \cr
   Total \tab 1831 \tab 1110 \tab 1265 \tab 106 \tab 22 \tab 10 \tab 11 \tab 16 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{rightpeo}},
\code{\link{vistholy}}
}
\concept{ISSP Religion}
\keyword{variable}
