% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{commun}
\alias{commun}
\title{Feelings about communism}
\source{
General Social Survey https://gss.norc.org
}
\description{
commun
}
\details{
Question Thinking about all the different kinds of governments in the world today, which of these statements comes closest to how you feel about Communism as a form of government? CODE ONE.
}
\section{Values}{

\itemize{
\item \code{1} worst kind
\item \code{2} bad, not worst
\item \code{3} ok for some
\item \code{4} good form
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5488/vshow}{GSS Data Explorer page for \code{commun}}.
}

\seealso{
Other War and International: 
\code{\link{commun10}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}

Other Communism: 
\code{\link{commun10}}
}
\concept{Communism}
\concept{War and International}
\keyword{variable}
