% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{commun10}
\alias{commun10}
\title{Feelings about communism 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
commun10
}
\details{
Question What about ten years ago, that is in 1972, which of these statements came closest to how you felt about Communism as a form of government?  CODE ONE.
}
\section{Values}{

\itemize{
\item \code{1} worst kind
\item \code{2} bad, not worst
\item \code{3} ok for some
\item \code{4} good form
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5489/vshow}{GSS Data Explorer page for \code{commun10}}.
}

\seealso{
Other War and International: 
\code{\link{commun}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}
}
\concept{War and International}
\keyword{variable}
