% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_379.R
\name{comorsci}
\alias{comorsci}
\title{Antarctica should be reserved or open}
\source{
General Social Survey https://gss.norc.org
}
\description{
comorsci
}
\details{
Question Some people think that Antarctica should be reserved primarily for scientific purposes. Others think it should be open for tourism, fishing, exploration for oil, and other commercial purposes. Please look at Card B23. Which statement best describes your view on this issue?
}
\section{Values}{

\itemize{
\item \code{1} strongly support reserving antarctica
\item \code{2} somewhat support reserving antarctica
\item \code{3} dont lean one way or the other on this issue
\item \code{4} somewhat support opening antarctica
\item \code{5} strongly support opening antarctica
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3437/vshow}{GSS Data Explorer page for \code{comorsci}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{extinct}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{nosun}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}}

Other Environment: 
\code{\link{extinct}}

Other Science: 
\code{\link{extinct}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{nosun}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}}
}
\concept{Environment}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
