% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{company}
\alias{company}
\title{Which company would r prefer to work for}
\source{
General Social Survey https://gss.norc.org
}
\description{
company
}
\details{
Question What kind of company would you be most likely to work for if these were your choices?
}
\section{Values}{

\itemize{
\item \code{1} a company in which the government owns the stock and appoints the management to run the company's operations
\item \code{2} a company whose stock is owned by the employees who appoint the management to run the company's operations
\item \code{3} a company whose stock is owned by outside investors who appoint the management to run the company's operations
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7227/vshow}{GSS Data Explorer page for \code{company}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a company in which the government owns the stock and appoints the management to run the company's operations \tab a company whose stock is owned by the employees who appoint the management to run the company's operations \tab a company whose stock is owned by outside investors who appoint the management to run the company's operations \tab don't know \tab no answer \tab Total \cr
   2018 \tab 900 \tab 123 \tab 981 \tab 250 \tab 56 \tab 38 \tab 2348 \cr
   Total \tab 900 \tab 123 \tab 981 \tab 250 \tab 56 \tab 38 \tab 2348 \cr
}
}

\keyword{variable}
