% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{compoth}
\alias{compoth}
\title{Does r use computer not at home or work}
\source{
General Social Survey https://gss.norc.org
}
\description{
compoth
}
\details{
Question A. Do you use a computer at some other place besides your home or workplace–say, at school, library, friend''s  home, or other location?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2579/vshow}{GSS Data Explorer page for \code{compoth}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2294 \tab 124 \tab 371 \tab 28 \tab 2817 \cr
   2002 \tab 855 \tab 482 \tab 1415 \tab 13 \tab 2765 \cr
   Total \tab 3149 \tab 606 \tab 1786 \tab 41 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compschl}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Information Society}
\keyword{variable}
