% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{compschl}
\alias{compschl}
\title{Does r use computer at school}
\source{
General Social Survey https://gss.norc.org
}
\description{
compschl
}
\details{
Question B. School?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2580/vshow}{GSS Data Explorer page for \code{compschl}}.
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}

Other Computers And Internet: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{mostloc}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
