% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{compwage}
\alias{compwage}
\title{Annual wages in previous year were higher than same jobs in other companies}
\source{
General Social Survey https://gss.norc.org
}
\description{
compwage
}
\details{
Question Do you believe your fixed annual wages in calendar year 2005 were higher or lower than those of employees with similar experience and job descriptions in other companies in your region? Please answer on a 1 to 5 scale.
}
\section{Values}{

\itemize{
\item \code{1} 1 - lower
\item \code{5} 5 - higher
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3554/vshow}{GSS Data Explorer page for \code{compwage}}.
}

\seealso{
Other Shared Capitalism: 
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
