% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{compwork}
\alias{compwork}
\title{Does r use computer at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
compwork
}
\details{
Question Do you use a computer at work–that is, at a workplace away from your home?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2566/vshow}{GSS Data Explorer page for \code{compwork}}.
}

\seealso{
Other Information Society: 
\code{\link{emhrh}},
\code{\link{emminh}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{perhrh}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwhrh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
