% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_571.R
\name{comsys}
\alias{comsys}
\title{Punishment for damaging a computer system}
\source{
General Social Survey https://gss.norc.org
}
\description{
comsys
}
\details{
Question Now I'm going to ask you about several offenses involving the use of a government computer system. For each case, indicate what you would consider to be the appropriate punishment when the offender is a military service member or government employee.
C. Intentionally damaging or destroying a computer system.
}
\section{Values}{

\itemize{
\item \verb{[1]} reprimand
\item \verb{[2]} dismissal from the military/fired by the government
\item \verb{[3]} one year in prison
\item \verb{[4]} 5-9 years in prison
\item \verb{[5]} 10-20 years in prison
\item \verb{[6]} life in prison
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5239/vshow}{GSS Data Explorer page for \code{comsys}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab reprimand \tab dismissal from the military/fired by the government \tab one year in prison \tab 5-9 years in prison \tab 10-20 years in prison \tab life in prison \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 23 \tab 287 \tab 174 \tab 377 \tab 332 \tab 152 \tab 57 \tab 17 \tab 2817 \cr
   Total \tab 1398 \tab 23 \tab 287 \tab 174 \tab 377 \tab 332 \tab 152 \tab 57 \tab 17 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{punenmy}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}}
}
\concept{National Security}
\keyword{variable}
