% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{comtype}
\alias{comtype}
\title{Type of community in which r lives}
\source{
General Social Survey https://gss.norc.org
}
\description{
comtype
}
\details{
Question Would you describe the place where you live as…
}
\section{Values}{

\itemize{
\item \code{1} big city
\item \code{2} suburbs, outskirts
\item \code{3} small town
\item \code{4} country village
\item \code{5} farm, cntry home
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4777/vshow}{GSS Data Explorer page for \code{comtype}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
