% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{conage}
\alias{conage}
\title{Conflict between young and old}
\source{
General Social Survey https://gss.norc.org
}
\description{
conage
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
G. Young people and older people?
}
\section{Values}{

\itemize{
\item \verb{[1]} very strong conflicts
\item \verb{[2]} strong conflicts
\item \verb{[3]} not very strong conflicts
\item \verb{[4]} there are no conflicts
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4312/vshow}{GSS Data Explorer page for \code{conage}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very strong conflicts \tab strong conflicts \tab not very strong conflicts \tab there are no conflicts \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 118 \tab 359 \tab 551 \tab 79 \tab 51 \tab 114 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 138 \tab 537 \tab 925 \tab 116 \tab 111 \tab 159 \tab 17 \tab 4032 \cr
   Total \tab 3574 \tab 256 \tab 896 \tab 1476 \tab 195 \tab 162 \tab 273 \tab 17 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
