% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{concong}
\alias{concong}
\title{Confidence in us congress}
\source{
General Social Survey https://gss.norc.org
}
\description{
concong
}
\details{
Question How much confidence do you have in. . .
A. U.S. Congress
}
\section{Values}{

\itemize{
\item \code{1} complete confidence
\item \code{2} a great deal of confidence
\item \code{3} some confidence
\item \code{4} very little confidence
\item \code{5} no confidence at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4639/vshow}{GSS Data Explorer page for \code{concong}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}

Other Confidence: 
\code{\link{conbiz}},
\code{\link{congovt}}

Other Federal Government: 
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{taxcheat}}
}
\concept{Confidence}
\concept{Federal Government}
\concept{ISSP Religion}
\keyword{variable}
