% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{concourt}
\alias{concourt}
\title{Confidence in courts and legal system}
\source{
General Social Survey https://gss.norc.org
}
\description{
concourt
}
\details{
Question How much confidence do you have in. . .
E. Courts and the legal system
}
\section{Values}{

\itemize{
\item \verb{[1]} complete confidence
\item \verb{[2]} a great deal of confidence
\item \verb{[3]} some confidence
\item \verb{[4]} very little confidence
\item \verb{[5]} no confidence at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4643/vshow}{GSS Data Explorer page for \code{concourt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab complete confidence \tab a great deal of confidence \tab some confidence \tab very little confidence \tab no confidence at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 75 \tab 254 \tab 608 \tab 277 \tab 81 \tab 27 \tab 37 \tab 1517 \cr
   1998 \tab 1548 \tab 38 \tab 209 \tab 572 \tab 294 \tab 104 \tab 30 \tab 37 \tab 2832 \cr
   2008 \tab 658 \tab 60 \tab 239 \tab 733 \tab 227 \tab 87 \tab 15 \tab 4 \tab 2023 \cr
   2018 \tab 1173 \tab 49 \tab 194 \tab 575 \tab 260 \tab 80 \tab 10 \tab 7 \tab 2348 \cr
   Total \tab 3537 \tab 222 \tab 896 \tab 2488 \tab 1058 \tab 352 \tab 82 \tab 85 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{polsgod}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
