% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_465.R
\name{conf2f}
\alias{conf2f}
\title{How many people r sees face to face}
\source{
General Social Survey https://gss.norc.org
}
\description{
conf2f
}
\details{
Question About how many of these people do you see face-to-face on a typical weekday?
}
\section{Values}{

\itemize{
\item \verb{[1]} all or almost all of them
\item \verb{[2]} most of them
\item \verb{[3]} about half of them
\item \verb{[4]} some of them
\item \verb{[5]} none or almost none of them
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7264/vshow}{GSS Data Explorer page for \code{conf2f}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all or almost all of them \tab most of them \tab about half of them \tab some of them \tab none or almost none of them \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 189 \tab 308 \tab 236 \tab 332 \tab 102 \tab 4 \tab 2 \tab 2348 \cr
   Total \tab 1175 \tab 189 \tab 308 \tab 236 \tab 332 \tab 102 \tab 4 \tab 2 \tab 2348 \cr
}
}

\keyword{variable}
