% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{confed}
\alias{confed}
\title{Confid. in exec branch of fed govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
confed
}
\details{
Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?
READ EACH ITEM; CODE ONE FOR EACH.
E. Executive branch of the federal government
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
   1996 \tab B/C \tab partial \cr
   1998 \tab B/C \tab partial \cr
   2000 \tab B/C \tab partial \cr
   2002 \tab B/C \tab partial \cr
   2004 \tab B/C \tab partial \cr
   2006 \tab B/C \tab partial \cr
   2008 \tab B/C \tab partial \cr
   2010 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab B/C \tab partial \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/452/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 439 \tab 29 \tab 275 \tab 6 \tab 755 \tab - \tab 1504 \cr
   1974 \tab - \tab 202 \tab 32 \tab 618 \tab 2 \tab 630 \tab - \tab 1484 \cr
   1975 \tab - \tab 198 \tab 38 \tab 439 \tab 2 \tab 813 \tab - \tab 1490 \cr
   1976 \tab - \tab 201 \tab 45 \tab 374 \tab 5 \tab 874 \tab - \tab 1499 \cr
   1977 \tab - \tab 426 \tab 48 \tab 221 \tab 5 \tab 830 \tab - \tab 1530 \cr
   1978 \tab - \tab 191 \tab 49 \tab 380 \tab 4 \tab 908 \tab - \tab 1532 \cr
   1980 \tab - \tab 177 \tab 49 \tab 502 \tab 3 \tab 737 \tab - \tab 1468 \cr
   1982 \tab - \tab 323 \tab 49 \tab 495 \tab 9 \tab 984 \tab - \tab 1860 \cr
   1983 \tab - \tab 206 \tab 52 \tab 471 \tab 2 \tab 868 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 181 \tab 23 \tab 280 \tab 12 \tab 493 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 302 \tab 41 \tab 346 \tab 2 \tab 779 \tab - \tab 1470 \cr
   1987 \tab - \tab 316 \tab 65 \tab 530 \tab 4 \tab 904 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 164 \tab 34 \tab 266 \tab 3 \tab 530 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 202 \tab 47 \tab 225 \tab 4 \tab 557 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 210 \tab 30 \tab 208 \tab - \tab 451 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 263 \tab 24 \tab 212 \tab 3 \tab 515 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 124 \tab 32 \tab 340 \tab 5 \tab 556 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 226 \tab 42 \tab 701 \tab 8 \tab 1034 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 194 \tab 68 \tab 800 \tab 1 \tab 862 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 265 \tab 60 \tab 671 \tab 6 \tab 909 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 249 \tab 82 \tab 632 \tab 9 \tab 924 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 243 \tab 19 \tab 193 \tab 2 \tab 455 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 179 \tab 12 \tab 271 \tab 2 \tab 412 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 282 \tab 44 \tab 789 \tab 4 \tab 870 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 133 \tab 32 \tab 552 \tab 1 \tab 647 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 227 \tab 26 \tab 499 \tab 5 \tab 620 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 196 \tab 28 \tab 493 \tab 3 \tab 618 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 199 \tab 29 \tab 731 \tab 5 \tab 723 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 238 \tab 36 \tab 827 \tab 4 \tab 851 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 182 \tab 25 \tab 692 \tab 5 \tab 659 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 336 \tab 2 \tab 1182 \tab - \tab 1140 \tab 12 \tab 4032 \cr
   2022 \tab 1173 \tab 243 \tab 30 \tab 1125 \tab 4 \tab 962 \tab 7 \tab 3544 \cr
   2024 \tab 1116 \tab 266 \tab 31 \tab 1003 \tab 3 \tab 876 \tab 14 \tab 3309 \cr
   Total \tab 24408 \tab 7783 \tab 1253 \tab 17343 \tab 133 \tab 24746 \tab 33 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confinan}},
\code{\link{conlabor}},
\code{\link{conpress}},
\code{\link{sathealt}}

Other Confidence: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confinan}},
\code{\link{conlabor}},
\code{\link{conpress}}

Other Split Ballots: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confinan}},
\code{\link{conlabor}},
\code{\link{conpress}},
\code{\link{satfam}},
\code{\link{satfrnd}},
\code{\link{sathealt}}
}
\concept{Confidence}
\concept{Core}
\concept{Federal Government}
\concept{Split Ballots}
\keyword{variable}
