% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{confedvac}
\alias{confedvac}
\title{Confidence in the federal government x}
\source{
General Social Survey https://gss.norc.org
}
\description{
confedvac
}
\details{
Question Here are some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?

The federal government ensuring the safety of vaccines to protect the public against serious diseases
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7580/vshow}{GSS Data Explorer page for \code{confedvac}}.
}

\keyword{variable}
