% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{congeth}
\alias{congeth}
\title{Should ethnic congressmen represent ethnic groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
congeth
}
\details{
Question Some people say that it is only fair that in a society like ours, people holding public office should have roughly the same racial or ethnic background as the population as a whole. So, for example, if 30pct of the people were Hispanic, then 30pct of the public officials should be Hispanic as well. Others say that in choosing people for public office they should be considered purely on the basis of ability without regard to their ethnic or racial background. Let me ask you about these ideas in several different areas.
A. What about people in Congress?  Should they have the same racial or ethnic background as their constituents, or should they be considered purely on the basis of their ability without regard to their racial or ethnic background?
}
\section{Values}{

\itemize{
\item \verb{[1]} consider ethnic background
\item \verb{[2]} ability, not background
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1627/vshow}{GSS Data Explorer page for \code{congeth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab consider ethnic background \tab ability, not background \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 131 \tab 1279 \tab 53 \tab 11 \tab 2992 \cr
   Total \tab 1518 \tab 131 \tab 1279 \tab 53 \tab 11 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{discaffy}},
\code{\link{ethhist}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
