% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{conghlp1}
\alias{conghlp1}
\title{How much congregation help out with ill}
\source{
General Social Survey https://gss.norc.org
}
\description{
conghlp1
}
\details{
Question If you were ill, how much would the people in your congregation help you out?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} some
\item \verb{[3]} a little
\item \verb{[4]} none
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2092/vshow}{GSS Data Explorer page for \code{conghlp1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab some \tab a little \tab none \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1797 \tab 372 \tab 261 \tab 110 \tab 63 \tab 65 \tab 164 \tab 2832 \cr
   Total \tab 1797 \tab 372 \tab 261 \tab 110 \tab 63 \tab 65 \tab 164 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{meditate}},
\code{\link{privpray}}
}
\concept{Religion and Culture}
\keyword{variable}
