% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{conghrm1}
\alias{conghrm1}
\title{How often congregation make too much demand}
\source{
General Social Survey https://gss.norc.org
}
\description{
conghrm1
}
\details{
Question How often do the people in your congregation make too many demands on you . . .
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} fairly often
\item \code{3} once in a while
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2094/vshow}{GSS Data Explorer page for \code{conghrm1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab fairly often \tab once in a while \tab never \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1797 \tab 23 \tab 57 \tab 253 \tab 542 \tab 4 \tab 156 \tab 2832 \cr
   Total \tab 1797 \tab 23 \tab 57 \tab 253 \tab 542 \tab 4 \tab 156 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{meditate}},
\code{\link{privpray}}
}
\concept{Religion and Culture}
\keyword{variable}
