% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{congovt}
\alias{congovt}
\title{Confidence in government departments}
\source{
General Social Survey https://gss.norc.org
}
\description{
congovt
}
\details{
Question How much confidence do you have in. . .
C. Government departments
}
\section{Values}{

\itemize{
\item \verb{[1]} complete confidence
\item \verb{[2]} a great deal of confidence
\item \verb{[3]} some confidence
\item \verb{[4]} very little confidence
\item \verb{[5]} no confidence at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4641/vshow}{GSS Data Explorer page for \code{congovt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab complete confidence \tab a great deal of confidence \tab some confidence \tab very little confidence \tab no confidence at all \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 56 \tab 163 \tab 626 \tab 342 \tab 89 \tab 36 \tab 47 \tab 1517 \cr
   Total \tab 158 \tab 56 \tab 163 \tab 626 \tab 342 \tab 89 \tab 36 \tab 47 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{govcheat}},
\code{\link{homosex1}},
\code{\link{taxcheat}}
}
\concept{ISSP Religion}
\keyword{variable}
