% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{conhlth}
\alias{conhlth}
\title{Confidence in health care system in us}
\source{
General Social Survey https://gss.norc.org
}
\description{
conhlth
}
\details{
Question In general, how much confidence do you have in the health care system in the United States?
}
\section{Values}{

\itemize{
\item \code{1} complete confidence
\item \code{2} a great deal of confidence
\item \code{3} some confidence
\item \code{4} very little confidence
\item \code{5} no confidence at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7789/vshow}{GSS Data Explorer page for \code{conhlth}}.
}

\keyword{variable}
