% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{conimm}
\alias{conimm}
\title{Conflict between natives and immigrants}
\source{
General Social Survey https://gss.norc.org
}
\description{
conimm
}
\details{
Question People born in America and people from other countries who have come to live in America?
}
\section{Values}{

\itemize{
\item \verb{[1]} very strong conflicts
\item \verb{[2]} strong conflicts
\item \verb{[3]} not very strong conflicts
\item \verb{[4]} there are no conflicts
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7511/vshow}{GSS Data Explorer page for \code{conimm}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very strong conflicts \tab strong conflicts \tab not very strong conflicts \tab there are no conflicts \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 244 \tab 854 \tab 558 \tab 44 \tab 126 \tab 160 \tab 17 \tab 4032 \cr
   Total \tab 2029 \tab 244 \tab 854 \tab 558 \tab 44 \tab 126 \tab 160 \tab 17 \tab 4032 \cr
}
}

\keyword{variable}
