% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{conlabory}
\alias{conlabory}
\title{Confidence in labor unions}
\source{
General Social Survey https://gss.norc.org
}
\description{
conlabory
}
\details{
Question I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?
READ EACH ITEM; CODE ONE FOR EACH.
F. Labor unions
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8113/vshow}{GSS Data Explorer page for \code{conlabory}}.
}

\seealso{
Other Core: 
\code{\link{marsamey}},
\code{\link{mobiledata}},
\code{\link{pawrkgrw}},
\code{\link{racdif5}}
}
\concept{Core}
\keyword{variable}
