% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{conpharvacy}
\alias{conpharvacy}
\title{Confidence in pharmaceutical companies y}
\source{
General Social Survey https://gss.norc.org
}
\description{
conpharvacy
}
\details{
Question Here are some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them?

Pharmaceutical companies working to create safe vaccines to protect the public against COVID-19
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2021 \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7579/vshow}{GSS Data Explorer page for \code{conpharvacy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a great deal \tab only some \tab hardly any \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2658 \tab 599 \tab 448 \tab 145 \tab 169 \tab 13 \tab 4032 \cr
   2022 \tab 2309 \tab 545 \tab 478 \tab 210 \tab - \tab 2 \tab 3544 \cr
   Total \tab 4967 \tab 1144 \tab 926 \tab 355 \tab 169 \tab 15 \tab 7576 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{confedvacy}}
}
\concept{GSS Next}
\keyword{variable}
