% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{conscrp}
\alias{conscrp}
\title{Importance of military service}
\source{
General Social Survey https://gss.norc.org
}
\description{
conscrp
}
\details{
Question On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:... To be willing to serve in the military at a time of need.
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8173/vshow}{GSS Data Explorer page for \code{conscrp}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{brneffrt}},
\code{\link{immcrime1}},
\code{\link{immideas1}},
\code{\link{immjobs1}},
\code{\link{immwlfare}}
}
\concept{ISSP National Identity}
\keyword{variable}
