% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{consoc}
\alias{consoc}
\title{Conflict between social top and bottom}
\source{
General Social Survey https://gss.norc.org
}
\description{
consoc
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
F. People at the top of society and people at the bottom?
}
\section{Values}{

\itemize{
\item \code{1} very strong conflicts
\item \code{2} strong conflicts
\item \code{3} not very strong conflicts
\item \code{4} there are no conflicts
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4311/vshow}{GSS Data Explorer page for \code{consoc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very strong conflicts \tab strong conflicts \tab not very strong conflicts \tab there are no conflicts \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 324 \tab 461 \tab 274 \tab 32 \tab 67 \tab 114 \tab 2817 \cr
   Total \tab 1545 \tab 324 \tab 461 \tab 274 \tab 32 \tab 67 \tab 114 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
