% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{contmslm}
\alias{contmslm}
\title{Have muslims positive contribution to country}
\source{
General Social Survey https://gss.norc.org
}
\description{
contmslm
}
\details{
Question Since the beginning of our country, people of many different races, nationalities, and religions have come here. and settled. As I name some of these groups please tell me if the group has made one of the most important positive contributions to this country, an important contribution, some contribution, or little positive contribution to this community.
L. Muslims?
}
\section{Values}{

\itemize{
\item \code{1} most important contribution
\item \code{2} important contribution
\item \code{3} some contribution
\item \code{4} little positive contribution
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2282/vshow}{GSS Data Explorer page for \code{contmslm}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
