% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{conurban}
\alias{conurban}
\title{Farmer's and city people in conflict?}
\source{
General Social Survey https://gss.norc.org
}
\description{
conurban
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
E. Farmers and city people?
}
\section{Values}{

\itemize{
\item \code{1} very strong conflict
\item \code{2} strong conflict
\item \code{3} not strong conflict
\item \code{4} no conflict
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4310/vshow}{GSS Data Explorer page for \code{conurban}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
