% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{conwlth}
\alias{conwlth}
\title{Conflict between poor and rich in usa?}
\source{
General Social Survey https://gss.norc.org
}
\description{
conwlth
}
\details{
Question In all countries, there are differences or conflicts between different social groups. In your opinion, in America, how much conflict is there between . . .
A. Poor people and rich people?
}
\section{Values}{

\itemize{
\item \verb{[1]} very strong conflicts
\item \verb{[2]} strong conflicts
\item \verb{[3]} not very strong conflicts
\item \verb{[4]} there are no conflicts
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4306/vshow}{GSS Data Explorer page for \code{conwlth}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very strong conflicts \tab strong conflicts \tab not very strong conflicts \tab there are no conflicts \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 270 \tab 657 \tab 468 \tab 54 \tab 75 \tab 40 \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 157 \tab 472 \tab 425 \tab 33 \tab 72 \tab 113 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 255 \tab 849 \tab 526 \tab 45 \tab 150 \tab 158 \tab 20 \tab 4032 \cr
   Total \tab 3829 \tab 682 \tab 1978 \tab 1419 \tab 132 \tab 297 \tab 311 \tab 20 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
