% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{copres10}
\alias{copres10}
\title{Partner's occupational prestige score (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
copres10
}
\details{
Question COPRES10 is based on the 2010 Census occupation classification. This standard prestige score is a simple mean value of ratings for each occupation category, converted to a scale of 0 (bottom) to 100 (top).
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6159/vshow}{GSS Data Explorer page for \code{copres10}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 18 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 53 \tab 54 \tab 55 \tab 59 \tab 60 \tab 61 \tab 63 \tab 64 \tab 65 \tab 66 \tab 69 \tab 71 \tab 72 \tab 73 \tab 30 \tab 52 \tab 56 \tab 57 \tab 62 \tab 75 \tab 80 \tab 17 \tab 58 \tab 68 \tab 74 \tab 16 \tab 70 \tab Total \cr
   2012 \tab 1824 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab 3 \tab 2 \tab 7 \tab 2 \tab 13 \tab 3 \tab 5 \tab 2 \tab 13 \tab 7 \tab 1 \tab 9 \tab 7 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 8 \tab 5 \tab 2 \tab 8 \tab 3 \tab 3 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2348 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 12 \tab 9 \tab 1 \tab 1 \tab 12 \tab 1 \tab 10 \tab 4 \tab 4 \tab 1 \tab 20 \tab 7 \tab 3 \tab 11 \tab 3 \tab 5 \tab - \tab 2 \tab 3 \tab 4 \tab 16 \tab 2 \tab 3 \tab 9 \tab 5 \tab 5 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab 5 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2628 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 7 \tab 8 \tab 2 \tab 3 \tab 10 \tab - \tab 18 \tab 10 \tab 7 \tab 2 \tab 26 \tab 5 \tab 4 \tab 14 \tab 7 \tab 2 \tab - \tab 5 \tab 1 \tab 5 \tab 11 \tab 3 \tab 6 \tab 8 \tab 9 \tab 5 \tab 8 \tab 4 \tab 1 \tab 2 \tab 5 \tab 7 \tab 2 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2867 \cr
   2018 \tab 2131 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 11 \tab 9 \tab - \tab 2 \tab 7 \tab - \tab 11 \tab 4 \tab 6 \tab 1 \tab 18 \tab 2 \tab 4 \tab 15 \tab 15 \tab 8 \tab 1 \tab 4 \tab 6 \tab 2 \tab 4 \tab 4 \tab 6 \tab 10 \tab 6 \tab 3 \tab 5 \tab 6 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab 10 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 2348 \cr
   2021 \tab 3705 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 5 \tab 15 \tab - \tab 5 \tab 14 \tab 3 \tab 24 \tab 8 \tab 7 \tab 1 \tab 31 \tab 8 \tab 11 \tab 19 \tab 16 \tab 4 \tab - \tab 2 \tab 10 \tab 9 \tab 9 \tab 4 \tab 7 \tab 12 \tab 12 \tab 4 \tab 7 \tab 6 \tab 4 \tab 4 \tab 5 \tab 8 \tab 1 \tab 3 \tab 8 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab - \tab 2 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 16180 \tab 3 \tab 4 \tab 8 \tab 12 \tab 6 \tab 37 \tab 48 \tab 6 \tab 13 \tab 50 \tab 6 \tab 76 \tab 29 \tab 29 \tab 7 \tab 108 \tab 29 \tab 23 \tab 68 \tab 48 \tab 21 \tab 2 \tab 16 \tab 21 \tab 22 \tab 48 \tab 18 \tab 24 \tab 47 \tab 35 \tab 20 \tab 24 \tab 22 \tab 10 \tab 15 \tab 12 \tab 24 \tab 10 \tab 8 \tab 30 \tab 7 \tab 3 \tab 6 \tab 2 \tab 4 \tab 4 \tab 4 \tab 9 \tab 6 \tab 2 \tab 5 \tab 3 \tab 6 \tab 3 \tab 4 \tab 1 \tab 8 \tab 2 \tab 5 \tab 17303 \cr
}
}

\concept{Core}
\keyword{variable}
