% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{copunish}
\alias{copunish}
\title{Punish draft resister's?}
\source{
General Social Survey https://gss.norc.org
}
\description{
copunish
}
\details{
Question Even though they are no longer drafted for military service, young men are still required by law to register for the draft when they become 18 years old. If a young man refuses to register for the draft, do you think he should be punished in any way?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/816/vshow}{GSS Data Explorer page for \code{copunish}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 956 \tab 446 \tab 50 \tab 21 \tab 1473 \cr
   Total \tab 0 \tab 956 \tab 446 \tab 50 \tab 21 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}
}
\concept{Core}
\keyword{variable}
