% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_41.R
\name{corel}
\alias{corel}
\title{Partner's religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
corel
}
\details{
Question What is your (PARTNER'S) religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} christian
\item \code{12} native american
\item \code{13} inter-nondenominational
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6169/vshow}{GSS Data Explorer page for \code{corel}}.
}

\seealso{
Other Core: 
\code{\link{spden16}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spoth16}},
\code{\link{spothjew}},
\code{\link{sprel16}}
}
\concept{Core}
\keyword{variable}
