% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{corrupt}
\alias{corrupt}
\title{Must be corrupt to get to top in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
corrupt
}
\details{
Question Do you agree or disagree:
C. To get all the way to the top in America today, you have to be corrupt.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4322/vshow}{GSS Data Explorer page for \code{corrupt}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\concept{United States}
\keyword{variable}
