% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{corrupt2}
\alias{corrupt2}
\title{Opinion of corruption by government administrators}
\source{
General Social Survey https://gss.norc.org
}
\description{
corrupt2
}
\details{
Question And in your opinion, about how many government administrators in America are involved in corruption?
}
\section{Values}{

\itemize{
\item \code{1} almost none
\item \code{2} a few
\item \code{3} some
\item \code{4} quite a lot
\item \code{5} almost all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4124/vshow}{GSS Data Explorer page for \code{corrupt2}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bribe}},
\code{\link{corrupt1}},
\code{\link{knowpols}}

Other Political: 
\code{\link{bribe}},
\code{\link{corrupt1}},
\code{\link{knowpols}}
}
\concept{ISSP Role of Gov't}
\concept{Political}
\keyword{variable}
