% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{corruptn}
\alias{corruptn}
\title{How widespread corruption is in pub service in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
corruptn
}
\details{
Question How widespread do you think corruption is in the public service in America?
}
\section{Values}{

\itemize{
\item \verb{[1]} hardly anyone is involved
\item \verb{[2]} a small number of people are involved
\item \verb{[3]} a moderate number of people are involved
\item \verb{[4]} a lot of people are involved
\item \verb{[5]} almost everyone is involved
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4959/vshow}{GSS Data Explorer page for \code{corruptn}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab hardly anyone is involved \tab a small number of people are involved \tab a moderate number of people are involved \tab a lot of people are involved \tab almost everyone is involved \tab can't choose \tab no answer \tab not available in this year \tab skipped on web \tab Total \cr
   2004 \tab 1340 \tab 18 \tab 418 \tab 573 \tab 309 \tab 115 \tab 32 \tab 7 \tab - \tab - \tab 2812 \cr
   2014 \tab 1274 \tab 9 \tab 270 \tab 449 \tab 353 \tab 98 \tab 79 \tab 6 \tab - \tab - \tab 2538 \cr
   2024 \tab 1666 \tab 35 \tab 265 \tab 444 \tab 564 \tab 168 \tab 142 \tab 8 \tab - \tab 17 \tab 3309 \cr
   Total \tab 4280 \tab 62 \tab 953 \tab 1466 \tab 1226 \tab 381 \tab 253 \tab 21 \tab 0 \tab 17 \tab 8659 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}

Other Political: 
\code{\link{fixmistk}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\concept{Political}
\keyword{variable}
