% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{corruptn}
\alias{corruptn}
\title{How widespread corruption is in pub service in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
corruptn
}
\details{
Question How widespread do you think corruption is in the public service in America?
}
\section{Values}{

\itemize{
\item \code{1} hardly anyone is involved
\item \code{2} a small number of people are involved
\item \code{3} a moderate number of people are involved
\item \code{4} a lot of people are involved
\item \code{5} almost everyone is involved
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4959/vshow}{GSS Data Explorer page for \code{corruptn}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
