% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_598.R
\name{cosei10}
\alias{cosei10}
\title{Partner's socioeconomic index (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
cosei10
}
\details{
Question COSEI10 is based on the 2010 Census occupation classification. It is estimated from 539 occupational categories, using COPRES105PLUS.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} i don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{cosei10educ}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}}
}
\concept{Occupation}
\keyword{variable}
