% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{cosei10inc}
\alias{cosei10inc}
\title{Percentage of $45k+ earners in coocc10 based on acs 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
cosei10inc
}
\details{
Question Percentage of those who earned $45,000 or more among those who had the same occupation as in COOCC10 in American Community Survey between 2010 and 2012 (25-64 years old only).This percentage is not from the GSS but it is used to calculate COSEI10.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} i don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6360/vshow}{GSS Data Explorer page for \code{cosei10inc}}.
}

\seealso{
Other Occupation: 
\code{\link{copres105plus}}
}
\concept{Occupation}
\keyword{variable}
