% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{courtsnv}
\alias{courtsnv}
\title{Courts dealing with criminals (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
courtsnv
}
\details{
Question In general, do you think the courts in this area deal too harshly or not harshly enough with criminals?
}
\section{Values}{

\itemize{
\item \code{1} too harshly
\item \code{2} not harshly enough
\item \code{3} about right (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7606/vshow}{GSS Data Explorer page for \code{courtsnv}}.
}

\keyword{variable}
