% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{courtsnv}
\alias{courtsnv}
\title{Courts dealing with criminals (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
courtsnv
}
\details{
Question In general, do you think the courts in this area deal too harshly or not harshly enough with criminals?
}
\section{Values}{

\itemize{
\item \verb{[1]} too harshly
\item \verb{[2]} not harshly enough
\item \verb{[3]} about right (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7606/vshow}{GSS Data Explorer page for \code{courtsnv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too harshly \tab not harshly enough \tab about right (phone mode only: volunteered) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2045 \tab 581 \tab 1244 \tab 46 \tab 18 \tab 3 \tab 95 \tab 4032 \cr
   2022 \tab 2655 \tab 242 \tab 608 \tab 5 \tab 4 \tab 2 \tab 28 \tab 3544 \cr
   Total \tab 4700 \tab 823 \tab 1852 \tab 51 \tab 22 \tab 5 \tab 123 \tab 7576 \cr
}
}

\keyword{variable}
