% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{courtsy}
\alias{courtsy}
\title{Courts dealing with criminals-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
courtsy
}
\details{
Question In general, do you think the courts in this area deal too harshly, or not harshly enough with criminals, or don't you have enough information about the courts to say?
}
\section{Values}{

\itemize{
\item \code{1} too harsh
\item \code{2} not harsh enough
\item \code{3} about right
\item \code{8} cant say
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/277/vshow}{GSS Data Explorer page for \code{courtsy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too harsh \tab not harsh enough \tab about right \tab can't choose \tab no answer \tab Total \cr
   1974 \tab 753 \tab 33 \tab 436 \tab 44 \tab 210 \tab 8 \tab 1484 \cr
   1982 \tab 923 \tab 57 \tab 689 \tab 42 \tab 138 \tab 11 \tab 1860 \cr
   Total \tab 1676 \tab 90 \tab 1125 \tab 86 \tab 348 \tab 19 \tab 3344 \cr
}
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
