% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{covemply}
\alias{covemply}
\title{Change in employment over covid pandemic}
\source{
General Social Survey https://gss.norc.org
}
\description{
covemply
}
\details{
Question Considering your work activity before the COVID-19 pandemic and at present, which of the following statements best describes your employment situation?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} r did not have job before pandemic and does not have job now
\item \code{2} r has the same job as before the pandemic
\item \code{3} r lost job due to the pandemic and now has a new job
\item \code{4} r lost job due to the pandemic and could not find a new job
\item \code{5} r did not have job before the pandemic and does have job now
\item \code{6} r changed or left job for reasons unrelated to the pandemic
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7953/vshow}{GSS Data Explorer page for \code{covemply}}.
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covid12}},
\code{\link{fluvax}},
\code{\link{malekids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
