% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_99.R
\name{cowork1}
\alias{cowork1}
\title{Person #1 is r's coworker}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowork1
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
F1. First Person is Co-Worker
}
\section{Values}{

\itemize{
\item \code{1} mentioned
\item \code{2} not mentioned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/900/vshow}{GSS Data Explorer page for \code{cowork1}}.
}

\seealso{
Other Social Networks: 
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}}
}
\concept{Social Networks}
\keyword{variable}
