% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{cowrkers}
\alias{cowrkers}
\title{Quality of rels among workers: r's firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkers
}
\details{
Question In general how would you describe relations in your work place…
B. between co-workers/colleagues?
}
\section{Values}{

\itemize{
\item \code{1} very good
\item \code{2} quite good
\item \code{3} neither good nor bad
\item \code{4} quite bad
\item \code{5} very bad
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4560/vshow}{GSS Data Explorer page for \code{cowrkers}}.
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
