% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{cowrkfrd}
\alias{cowrkfrd}
\title{Number of close friends at work place}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkfrd
}
\details{
Question Now we would like to ask you about people you know, other than your family and relatives. The first question is about people at your work place. Thinking about people at your work place, how many of them are close friends of yours?
}
\section{Values}{

\itemize{
\item \code{0} no close friends at work
\item \code{96} more than a hundred
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4202/vshow}{GSS Data Explorer page for \code{cowrkfrd}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab no close friends at work \tab 1 \tab 10 \tab 12 \tab 15 \tab 16 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 25 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2002 \tab 206 \tab 121 \tab 25 \tab 3 \tab 6 \tab 1 \tab 1 \tab 146 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 84 \tab 2 \tab 57 \tab 46 \tab 33 \tab 5 \tab 9 \tab 4 \tab 1 \tab 6 \tab 2765 \cr
   Total \tab 2002 \tab 206 \tab 121 \tab 25 \tab 3 \tab 6 \tab 1 \tab 1 \tab 146 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 84 \tab 2 \tab 57 \tab 46 \tab 33 \tab 5 \tab 9 \tab 4 \tab 1 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
