% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{cowrkhrd}
\alias{cowrkhrd}
\title{How hard to say how people work at workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkhrd
}
\details{
Question At your workplace, how hard would you say that people work with 0 meaning not at all hard and 10 meaning very hard?
}
\section{Values}{

\itemize{
\item \code{0} not at all hard
\item \code{10} very hard
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2864/vshow}{GSS Data Explorer page for \code{cowrkhrd}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{deptperf}},
\code{\link{donothng}},
\code{\link{extr2001}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{ownstock}},
\code{\link{seecowrk}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{workfor}},
\code{\link{workfor1}},
\code{\link{yearval}},
\code{\link{yearval1}}
}
\concept{Employee Compensation}
\keyword{variable}
