% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{cowrkrel}
\alias{cowrkrel}
\title{Relations between coworkers}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkrel
}
\details{
Question In general how would you describe relations in your work place between coworkers/colleagues?
}
\section{Values}{

\itemize{
\item \code{1} very
\item \code{2} quite good
\item \code{3} neither good nor bad
\item \code{4} quite bad
\item \code{5} very bad
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1367/vshow}{GSS Data Explorer page for \code{cowrkrel}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very \tab quite good \tab neither good nor bad \tab quite bad \tab very bad \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 378 \tab 342 \tab 122 \tab 19 \tab 3 \tab 16 \tab 32 \tab 1517 \cr
   Total \tab 605 \tab 378 \tab 342 \tab 122 \tab 19 \tab 3 \tab 16 \tab 32 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{jobvshme}},
\code{\link{schoolng}},
\code{\link{unmanrel}},
\code{\link{unvote}}
}
\concept{Work Organizations}
\keyword{variable}
