% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_04.R
\name{cowrkslf}
\alias{cowrkslf}
\title{Partner self-emp. or works for somebody else}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkslf
}
\details{
Question A. What kind of work (does/did) your (PARTNER) normally do? That is, what (is/was) (his/her) job called?
OCCUPATION: (See REMARKS)
B. What (does/did) (PARTNER) actually do in that job? Tell me, what (are/were) some of (his/her) main duties?
C. What kind of place (does/did) (PARTNER) work for?
INDUSTRY: (See REMARKS)
D. What (do/did) they (make/do)?
E. (Is/Was) (he/she) self-employed or (does/did) (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6153/vshow}{GSS Data Explorer page for \code{cowrkslf}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab self-employed \tab someone else \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 1815 \tab 11 \tab 144 \tab 2 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 2341 \tab 25 \tab 167 \tab 2 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 2622 \tab 27 \tab 214 \tab 1 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 2127 \tab 21 \tab 198 \tab - \tab 2 \tab - \tab 2348 \cr
   2021 \tab 3675 \tab 41 \tab 303 \tab 1 \tab 8 \tab 4 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 16124 \tab 125 \tab 1026 \tab 6 \tab 18 \tab 4 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{coevwork}},
\code{\link{cohrs1}},
\code{\link{cowrksta}},
\code{\link{divorce}},
\code{\link{spevwork}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spwrksta}},
\code{\link{widowed}}
}
\concept{Core}
\keyword{variable}
