% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_664.R
\name{cowrkslf2}
\alias{cowrkslf2}
\title{Partner self-employed, with employees, or work for someone else}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrkslf2
}
\details{
Question Does/did your (husband/wife/spouse/partner) work without employees or does/did they have employees?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, employees
\item \verb{[2]} no, no employees
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8010/vshow}{GSS Data Explorer page for \code{cowrkslf2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, employees \tab no, no employees \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3242 \tab 24 \tab 268 \tab 4 \tab 4 \tab 2 \tab 3544 \cr
   Total \tab 9622 \tab 24 \tab 268 \tab 4 \tab 4 \tab 2 \tab 9924 \cr
}
}

\keyword{variable}
