% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_04.R
\name{cowrksta}
\alias{cowrksta}
\title{Partner's labor force status}
\source{
General Social Survey https://gss.norc.org
}
\description{
cowrksta
}
\details{
Question Last week was your partner working full time, part time, going to school, keeping house, or what?
}
\section{Values}{

\itemize{
\item \code{1} working full time
\item \code{2} working part time
\item \code{3} with a job, but not at work because of temporary illness, vacation, strike
\item \code{4} unemployed, laid off, looking for work
\item \code{5} retired
\item \code{6} in school
\item \code{7} keeping house
\item \code{8} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6152/vshow}{GSS Data Explorer page for \code{cowrksta}}.
}

\seealso{
Other Core: 
\code{\link{coevwork}},
\code{\link{cohrs1}},
\code{\link{cowrkslf}},
\code{\link{divorce}},
\code{\link{spevwork}},
\code{\link{sphrs1}},
\code{\link{sphrs2}},
\code{\link{spwrksta}},
\code{\link{widowed}}
}
\concept{Core}
\keyword{variable}
