% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{creation}
\alias{creation}
\title{View about origin and development of man}
\source{
General Social Survey https://gss.norc.org
}
\description{
creation
}
\details{
Question After I read off three statements, please tell me which one comes closest to your views about the origin and development of man.
}
\section{Values}{

\itemize{
\item \code{1} god created man
\item \code{2} man has evolved
\item \code{3} man has evolved, but god guided
\item \code{4} other
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4972/vshow}{GSS Data Explorer page for \code{creation}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab god created man \tab man has evolved \tab man has evolved, but god guided \tab other \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 608 \tab 170 \tab 598 \tab 52 \tab 26 \tab 18 \tab 2812 \cr
   Total \tab 1340 \tab 608 \tab 170 \tab 598 \tab 52 \tab 26 \tab 18 \tab 2812 \cr
}
}

\seealso{
Other Misc II: 
\code{\link{scitesty}}
}
\concept{Misc II}
\keyword{variable}
