% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{creation}
\alias{creation}
\title{View about origin and development of man}
\source{
General Social Survey https://gss.norc.org
}
\description{
creation
}
\details{
Question After I read off three statements, please tell me which one comes closest to your views about the origin and development of man.
}
\section{Values}{

\itemize{
\item \code{1} god created man
\item \code{2} man has evolved
\item \code{3} man has evolved, but god guided
\item \code{4} other
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4972/vshow}{GSS Data Explorer page for \code{creation}}.
}

\seealso{
Other Misc II: 
\code{\link{scitesty}}

Other Religion: 
\code{\link{scitesty}}
}
\concept{Misc II}
\concept{Religion}
\keyword{variable}
