% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{credchoice}
\alias{credchoice}
\title{R's credit score has to do with choices or things out of your control}
\source{
General Social Survey https://gss.norc.org
}
\description{
credchoice
}
\details{
Question Would you say your current credit score has more to do with choices you’ve made or things out of your control?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} entirely choices you've made
\item \verb{[2]} mostly choices you've made
\item \verb{[3]} both choices you've made and things out of your control
\item \verb{[4]} mostly things out of your control
\item \verb{[5]} entirely things out of your control
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8264/vshow}{GSS Data Explorer page for \code{credchoice}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab entirely choices you've made \tab mostly choices you've made \tab both choices you've made and things out of your control \tab mostly things out of your control \tab entirely things out of your control \tab iap \tab skipped on web \tab Total \cr
   2024 \tab - \tab 352 \tab 331 \tab 304 \tab 57 \tab 20 \tab 2244 \tab 1 \tab 3309 \cr
   Total \tab 0 \tab 352 \tab 331 \tab 304 \tab 57 \tab 20 \tab 2244 \tab 1 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}},
\code{\link{petsseq}}

Other Income: 
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{credrep}}
}
\concept{GSS Next}
\concept{Income}
\keyword{variable}
