% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_688.R
\name{credrep}
\alias{credrep}
\title{R credit report}
\source{
General Social Survey https://gss.norc.org
}
\description{
credrep
}
\details{
Question In general, would you say that your credit record is poor, fair, good, very good, or excellent?
NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} poor
\item \verb{[2]} fair
\item \verb{[3]} good
\item \verb{[4]} very good
\item \verb{[5]} excellent
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8290/vshow}{GSS Data Explorer page for \code{credrep}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab poor \tab fair \tab good \tab very good \tab excellent \tab can't choose \tab iap \tab Total \cr
   2024 \tab - \tab 97 \tab 141 \tab 172 \tab 161 \tab 465 \tab 29 \tab 2244 \tab 3309 \cr
   Total \tab 0 \tab 97 \tab 141 \tab 172 \tab 161 \tab 465 \tab 29 \tab 2244 \tab 3309 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{abhelp1}},
\code{\link{abhelp2}},
\code{\link{abhelp3}},
\code{\link{abhelp4}},
\code{\link{abmoral}},
\code{\link{advfront}},
\code{\link{biokids}},
\code{\link{blkdsrv}},
\code{\link{blkmblty}},
\code{\link{blktry}},
\code{\link{brv16}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16dadnum}},
\code{\link{brv16filnum}},
\code{\link{brv16frndnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16grand}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16oth}},
\code{\link{brv16othnum}},
\code{\link{brv16par}},
\code{\link{brv16partnum}},
\code{\link{brv16sib}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{brv16sp}},
\code{\link{brv16spnum}},
\code{\link{brv5}},
\code{\link{brv5bronum}},
\code{\link{brv5child}},
\code{\link{brv5chinum}},
\code{\link{brv5cowknum}},
\code{\link{brv5cuznum}},
\code{\link{brv5dadnum}},
\code{\link{brv5daunum}},
\code{\link{brv5filnum}},
\code{\link{brv5frndnum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5grand}},
\code{\link{brv5milnum}},
\code{\link{brv5momnum}},
\code{\link{brv5oth}},
\code{\link{brv5othnum}},
\code{\link{brv5par}},
\code{\link{brv5partnum}},
\code{\link{brv5sib}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}},
\code{\link{brv5sonnum}},
\code{\link{brv5sp}},
\code{\link{brv5spnum}},
\code{\link{chldidel_next}},
\code{\link{chldpref_next}},
\code{\link{confedvacy}},
\code{\link{conpharvacy}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{famgen_next}},
\code{\link{femself}},
\code{\link{firstkidsex}},
\code{\link{fluvax}},
\code{\link{fund_next}},
\code{\link{gestate}},
\code{\link{happy_next}},
\code{\link{hhtype1_next}},
\code{\link{knwsexor}},
\code{\link{malekids}},
\code{\link{mascself}},
\code{\link{nexteligible}},
\code{\link{nextgen}},
\code{\link{nonbinkids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{perla_self}},
\code{\link{petsseq}},
\code{\link{scientbe}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{toofast}},
\code{\link{topprob1_next}},
\code{\link{topprob2_crim}},
\code{\link{topprob2_econ}},
\code{\link{topprob2_educ}},
\code{\link{topprob2_envr}},
\code{\link{topprob2_hlth}},
\code{\link{topprob2_immg}},
\code{\link{topprob2_none}},
\code{\link{topprob2_povt}},
\code{\link{topprob2_terr}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}

Other Income: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{credam}},
\code{\link{credcar}},
\code{\link{credchoice}},
\code{\link{credemp}},
\code{\link{credland}},
\code{\link{earndes}},
\code{\link{earnmore}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{fairearn}},
\code{\link{famfinan}},
\code{\link{finalter}},
\code{\link{finan1}},
\code{\link{finan2}},
\code{\link{finan3}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{finrela}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{impfinan}},
\code{\link{incdef}},
\code{\link{inclstjb}},
\code{\link{incneed}},
\code{\link{incom16}},
\code{\link{income}},
\code{\link{income06}},
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{justpay}},
\code{\link{minfood}},
\code{\link{minfour}},
\code{\link{mininc}},
\code{\link{minthree}},
\code{\link{othpay}},
\code{\link{owninc}},
\code{\link{pastpay}},
\code{\link{paychild}},
\code{\link{paychnge}},
\code{\link{paydowel}},
\code{\link{payedtrn}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{payfam1}},
\code{\link{payhard}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{povline}},
\code{\link{realinc}},
\code{\link{realrinc}},
\code{\link{rincblls}},
\code{\link{rincom06}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}},
\code{\link{rincom98}},
\code{\link{rincome}},
\code{\link{salsatfn}},
\code{\link{satfin}},
\code{\link{welfare1}},
\code{\link{welfare2}},
\code{\link{welfare3}},
\code{\link{welfare4}},
\code{\link{welfare5}},
\code{\link{welfare6}}
}
\concept{GSS Next}
\concept{Income}
\keyword{variable}
