% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{crelesse}
\alias{crelesse}
\title{Emotion work less carefully past 4 wks}
\source{
General Social Survey https://gss.norc.org
}
\description{
crelesse
}
\details{
Question Didn’t do work or other activities as carefully as usual?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2682/vshow}{GSS Data Explorer page for \code{crelesse}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 221 \tab 1186 \tab 12 \tab 2817 \cr
   Total \tab 1398 \tab 221 \tab 1186 \tab 12 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{didlesse}},
\code{\link{didlessp}},
\code{\link{limitedp}},
\code{\link{modact}},
\code{\link{stairs}}
}
\concept{Health Status}
\keyword{variable}
