% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{crimdown}
\alias{crimdown}
\title{What if more restrictions reduce crime}
\source{
General Social Survey https://gss.norc.org
}
\description{
crimdown
}
\details{
Question Suppose research proves that more restrictions on handguns would reduce violent crime. Which of these two reactions would be closer to your position?
}
\section{Values}{

\itemize{
\item \verb{[1]} would change mind and support more restrictions
\item \verb{[2]} would still oppose more restrictions
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3060/vshow}{GSS Data Explorer page for \code{crimdown}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab would change mind and support more restrictions \tab would still oppose more restrictions \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2479 \tab 84 \tab 239 \tab 8 \tab 2 \tab 2812 \cr
   Total \tab 2479 \tab 84 \tab 239 \tab 8 \tab 2 \tab 2812 \cr
}
}

\seealso{
Other Guns: 
\code{\link{crimup}},
\code{\link{hguncrim}},
\code{\link{hgunlaw}}
}
\concept{Guns}
\keyword{variable}
