% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{crimimp}
\alias{crimimp}
\title{Importance of crime issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
crimimp
}
\details{
Question How important is the crime issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \code{1} one of most imp
\item \code{2} important
\item \code{3} not very imp
\item \code{4} not important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/278/vshow}{GSS Data Explorer page for \code{crimimp}}.
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{crimfirm}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{guninfo}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
