% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{cshutyp06}
\alias{cshutyp06}
\title{During march-may 2006, was r resident in household or institution}
\source{
General Social Survey https://gss.norc.org
}
\description{
cshutyp06
}
\details{
Question During March through May 2006, were you a resident in a household – such as living in a house, apartment or similar residence - or did you live elsewhere such as living in an institution like a college dorm, military barrack, work camp, shelter, residential treatment center, long-term medical care facility or nursing home?
}
\section{Values}{

\itemize{
\item \code{1} living in household
\item \code{2} living in institution
\item \code{3} homeless (volunteered)
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5589/vshow}{GSS Data Explorer page for \code{cshutyp06}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab living in household \tab living in institution \tab homeless (volunteered) \tab other \tab no answer \tab Total \cr
   2008 \tab - \tab 1994 \tab 21 \tab 6 \tab 1 \tab 1 \tab 2023 \cr
   2010 \tab - \tab 1991 \tab 43 \tab 2 \tab 7 \tab 1 \tab 2044 \cr
   Total \tab 0 \tab 3985 \tab 64 \tab 8 \tab 8 \tab 2 \tab 4067 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
