% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{cshutyp10}
\alias{cshutyp10}
\title{During april-june 2010, was r resident in household or institution}
\source{
General Social Survey https://gss.norc.org
}
\description{
cshutyp10
}
\details{
Question During April through June 2010, were you a resident in a household – such as living in a house, apartment or similar residence - or did you live elsewhere such as living in an institution like a college dorm, military barrack, work camp, shelter, residential treatment center, long-term medical care facility or nursing home?
}
\section{Values}{

\itemize{
\item \verb{[1]} living in household
\item \verb{[2]} living in institution
\item \verb{[3]} homeless (volunteered)
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5591/vshow}{GSS Data Explorer page for \code{cshutyp10}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab living in household \tab living in institution \tab homeless (volunteered) \tab other \tab no answer \tab Total \cr
   2012 \tab - \tab 1944 \tab 19 \tab 4 \tab 4 \tab 3 \tab 1974 \cr
   2014 \tab - \tab 2490 \tab 39 \tab 5 \tab 2 \tab 2 \tab 2538 \cr
   Total \tab 0 \tab 4434 \tab 58 \tab 9 \tab 6 \tab 5 \tab 4512 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp08}},
\code{\link{cshutyp12}}
}
\concept{Sampling and Weights}
\keyword{variable}
