% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{cshutyp10}
\alias{cshutyp10}
\title{During april-june 2010, was r resident in household or institution}
\source{
General Social Survey https://gss.norc.org
}
\description{
cshutyp10
}
\details{
Question During April through June 2010, were you a resident in a household – such as living in a house, apartment or similar residence - or did you live elsewhere such as living in an institution like a college dorm, military barrack, work camp, shelter, residential treatment center, long-term medical care facility or nursing home?
}
\section{Values}{

\itemize{
\item \code{1} living in household
\item \code{2} living in institution
\item \code{3} homeless (volunteered)
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5591/vshow}{GSS Data Explorer page for \code{cshutyp10}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp08}},
\code{\link{cshutyp12}}
}
\concept{Sampling and Weights}
\keyword{variable}
