% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{curempyr}
\alias{curempyr}
\title{How many yrs r worked for current employer}
\source{
General Social Survey https://gss.norc.org
}
\description{
curempyr
}
\details{
Question How many years have you worked for your current employer?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3873/vshow}{GSS Data Explorer page for \code{curempyr}}.
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{autonojb}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{natdeg}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{typevisa}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work3yrs}}
}
\concept{Global Economics}
\keyword{variable}
