% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_85.R
\name{cutspdr}
\alias{cutspdr}
\title{Cut govt spending--r's position}
\source{
General Social Survey https://gss.norc.org
}
\description{
cutspdr
}
\details{
Question Some people think the government should provide fewer services, even in areas such as health and education, in order to reduce spending. Other people feel it is important for the government to continue the services it now provides even if it means no reduction in spending. ENTER CODE IN BOXES.
A. Where would you place yourself on this scale, or haven't you thought much about this?
}
\section{Values}{

\itemize{
\item \verb{[0]} havent thought much
\item \verb{[1]} cut spending
\item \verb{[7]} maintain services
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/763/vshow}{GSS Data Explorer page for \code{cutspdr}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab havent thought much \tab cut spending \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab maintain services \tab don't know \tab no answer \tab Total \cr
   1980 \tab - \tab 121 \tab 121 \tab 123 \tab 179 \tab 271 \tab 195 \tab 158 \tab 261 \tab 35 \tab 4 \tab 1468 \cr
   Total \tab 0 \tab 121 \tab 121 \tab 123 \tab 179 \tab 271 \tab 195 \tab 158 \tab 261 \tab 35 \tab 4 \tab 1468 \cr
}
}

\seealso{
Other Core: 
\code{\link{cutspdfg}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imprelax}},
\code{\link{impwork}}
}
\concept{Core}
\keyword{variable}
