% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_656.R
\name{cvdtour}
\alias{cvdtour}
\title{Change in online tours since first year of covid}
\source{
General Social Survey https://gss.norc.org
}
\description{
cvdtour
}
\details{
Question (Compared to the first year of the COVID-19 pandemic (March 2020 to March 2021), are you now...) visiting or touring online art museums, galleries, or exhibits more often, at the same rate, less often, or do you not visit or tour online art museums at all?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \code{1} more often
\item \code{2} at the same rate
\item \code{3} less often
\item \code{4} r not attending these activities
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7884/vshow}{GSS Data Explorer page for \code{cvdtour}}.
}

\seealso{
Other NEA: 
\code{\link{cvdarmus}},
\code{\link{cvdarpo}},
\code{\link{cvdclass}},
\code{\link{cvdpod}},
\code{\link{cvdstmus}},
\code{\link{cvdstpo}}
}
\concept{NEA}
\keyword{variable}
