% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{daily}
\alias{daily}
\title{What is your daily wage before tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
daily
}
\details{
Question IF HOURLY WAGE:
B. What was your daily wage, before any taxes or deductions?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2162/vshow}{GSS Data Explorer page for \code{daily}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 100 \tab 125 \tab 127 \tab 128 \tab 175 \tab 35 \tab 400 \tab 45 \tab 50 \tab 60 \tab 65 \tab 70 \tab no answer \tab Total \cr
   1998 \tab 2785 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 33 \tab 2832 \cr
   Total \tab 2785 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 33 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{hourly}},
\code{\link{monthly}},
\code{\link{trusting}},
\code{\link{weekly}},
\code{\link{yearly}}
}
\concept{Jobs}
\keyword{variable}
