% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{dailywrk}
\alias{dailywrk}
\title{How r's daily work is organized}
\source{
General Social Survey https://gss.norc.org
}
\description{
dailywrk
}
\details{
Question Which of the following statements best describes how your daily work is organized?
}
\section{Values}{

\itemize{
\item \code{1} i am free to decide how my daily work is organized
\item \code{2} i can decide within certain limits
\item \code{3} i am not free to decide how my daily work is organized
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4612/vshow}{GSS Data Explorer page for \code{dailywrk}}.
}

\seealso{
Other Work Orientation: 
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}

Other Job: 
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Job}
\concept{Work Orientation}
\keyword{variable}
