% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{databank}
\alias{databank}
\title{Computer data threat to individual privacy}
\source{
General Social Survey https://gss.norc.org
}
\description{
databank
}
\details{
Question The federal government has a lot of different pieces of information about people which computers can bring together very quickly. Is this a very serious threat to individual privacy, a fairly serious threat, not a serious threat, or not a threat at all to individual privacy?
}
\section{Values}{

\itemize{
\item \code{1} very serious threat
\item \code{2} fairly serious
\item \code{3} not serious
\item \code{4} not a threat at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4003/vshow}{GSS Data Explorer page for \code{databank}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{progtax}},
\code{\link{verdict}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
