% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{dataprot}
\alias{dataprot}
\title{Responsibility for internet data protection}
\source{
General Social Survey https://gss.norc.org
}
\description{
dataprot
}
\details{
Question Who should be mostly responsible for protecting personal information when using the internet?
}
\section{Values}{

\itemize{
\item \verb{[1]} online companies - as they need to ensure they process personal information safely
\item \verb{[2]} people - as they need to take care of their personal information
\item \verb{[3]} public authorities - as they need to ensure that personal information is protected
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8103/vshow}{GSS Data Explorer page for \code{dataprot}}.

Counts by year:\tabular{llllllllll}{
   year \tab not available in this year \tab online companies - as they need to ensure they process personal information safely \tab people - as they need to take care of their personal information \tab public authorities - as they need to ensure that personal information is protected \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 542 \tab 593 \tab 212 \tab 271 \tab 1673 \tab 4 \tab 14 \tab 3309 \cr
   Total \tab 0 \tab 542 \tab 593 \tab 212 \tab 271 \tab 1673 \tab 4 \tab 14 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{modeorg}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
