% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{dataprot}
\alias{dataprot}
\title{Responsibility for internet data protection}
\source{
General Social Survey https://gss.norc.org
}
\description{
dataprot
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} online companies - as they need to ensure they process personal information safely
\item \code{2} people - as they need to take care of their personal information
\item \code{3} public authorities - as they need to ensure that personal information is protected
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
