% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{dataprot}
\alias{dataprot}
\title{Responsibility for internet data protection}
\source{
General Social Survey https://gss.norc.org
}
\description{
dataprot
}
\details{
Question Who should be mostly responsible for protecting personal information when using the internet?
}
\section{Values}{

\itemize{
\item \code{1} online companies - as they need to ensure they process personal information safely
\item \code{2} people - as they need to take care of their personal information
\item \code{3} public authorities - as they need to ensure that personal information is protected
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8103/vshow}{GSS Data Explorer page for \code{dataprot}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{classtech}},
\code{\link{eductech}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{trcong}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
