% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{datefrst}
\alias{datefrst}
\title{Month and yr began first job}
\source{
General Social Survey https://gss.norc.org
}
\description{
datefrst
}
\details{
Question Respondent's First Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{datesch}},
\code{\link{indfirst}},
\code{\link{neverwk}},
\code{\link{occfirst}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{wrkslfst}}
}
\concept{Family Mobility}
\keyword{variable}
