% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_478.R
\name{daycare5}
\alias{daycare5}
\title{Relative ok for kid ( 3yrs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
daycare5
}
\details{
Question Think of a child under 3 years old whose parents both have full-time jobs. How suitable do you think each of these child care arrangements would be for the child?
E. A relative?
}
\section{Values}{

\itemize{
\item \code{1} very suitable
\item \code{2} somewhat suitable
\item \code{3} not very suitable
\item \code{4} not at all suitable
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4361/vshow}{GSS Data Explorer page for \code{daycare5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very suitable \tab somewhat suitable \tab not very suitable \tab not at all suitable \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 703 \tab 479 \tab 89 \tab 44 \tab 59 \tab 40 \tab 1481 \cr
   Total \tab 67 \tab 703 \tab 479 \tab 89 \tab 44 \tab 59 \tab 40 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{ilikejob}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
