% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{daynight}
\alias{daynight}
\title{When did most errors occur}
\source{
General Social Survey https://gss.norc.org
}
\description{
daynight
}
\details{
Question Please look at Card I-3. Day-night rhythms dramatically affect our bodies. Probably no body system is more  influenced than the nervous system. The figure on Card I-3 illustrates the number of errors made by shift workers in different portions of the 24-hour cycle. Based on the data illustrated in the figure, during which of these time periods did the most errors occur?
}
\section{Values}{

\itemize{
\item \verb{[1]} 2 a.m. to 4 a.m.
\item \verb{[2]} 8 a.m. to 10 a.m.
\item \verb{[3]} 12 p.m. to 2 p.m.
\item \verb{[4]} 2 p.m. to 4 p.m.
\item \verb{[5]} 8 p.m. to 10 p.m.
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3482/vshow}{GSS Data Explorer page for \code{daynight}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 2 a.m. to 4 a.m. \tab 8 a.m. to 10 a.m. \tab 12 p.m. to 2 p.m. \tab 2 p.m. to 4 p.m. \tab 8 p.m. to 10 p.m. \tab don't know \tab no answer \tab Total \cr
   2008 \tab 709 \tab 999 \tab 52 \tab 54 \tab 46 \tab 30 \tab 111 \tab 22 \tab 2023 \cr
   Total \tab 709 \tab 999 \tab 52 \tab 54 \tab 46 \tab 30 \tab 111 \tab 22 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
