% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{death16}
\alias{death16}
\title{Relatives died since age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
death16
}
\details{
Question B. Number of relatives dying since respondent 16 years old and prior to the last five years.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/747/vshow}{GSS Data Explorer page for \code{death16}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab Total \cr
   1978 \tab - \tab 766 \tab 249 \tab 173 \tab 112 \tab 90 \tab 65 \tab 49 \tab 20 \tab 7 \tab 1 \tab 1532 \cr
   1980 \tab - \tab 722 \tab 227 \tab 175 \tab 122 \tab 88 \tab 67 \tab 42 \tab 15 \tab 9 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 807 \tab 273 \tab 165 \tab 117 \tab 95 \tab 81 \tab 42 \tab 16 \tab 2 \tab 1 \tab 1599 \cr
   1984 \tab - \tab 762 \tab 236 \tab 148 \tab 117 \tab 79 \tab 58 \tab 42 \tab 20 \tab 8 \tab 3 \tab 1473 \cr
   1986 \tab - \tab 701 \tab 258 \tab 142 \tab 105 \tab 98 \tab 81 \tab 55 \tab 28 \tab 2 \tab - \tab 1470 \cr
   1987 \tab - \tab 930 \tab 306 \tab 200 \tab 127 \tab 97 \tab 80 \tab 59 \tab 16 \tab 3 \tab 1 \tab 1819 \cr
   1988 \tab 484 \tab 503 \tab 155 \tab 105 \tab 71 \tab 60 \tab 44 \tab 36 \tab 19 \tab 4 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 533 \tab 178 \tab 89 \tab 85 \tab 59 \tab 48 \tab 32 \tab 7 \tab 2 \tab 2 \tab 1537 \cr
   1990 \tab 473 \tab 456 \tab 155 \tab 87 \tab 60 \tab 60 \tab 42 \tab 24 \tab 13 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 520 \tab 176 \tab 102 \tab 63 \tab 64 \tab 53 \tab 20 \tab 15 \tab 3 \tab 1 \tab 1517 \cr
   1993 \tab 549 \tab 555 \tab 180 \tab 105 \tab 63 \tab 62 \tab 41 \tab 31 \tab 17 \tab 2 \tab 1 \tab 1606 \cr
   1994 \tab 2481 \tab 257 \tab 92 \tab 48 \tab 36 \tab 31 \tab 26 \tab 10 \tab 7 \tab 3 \tab 1 \tab 2992 \cr
   Total \tab 4989 \tab 7512 \tab 2485 \tab 1539 \tab 1078 \tab 883 \tab 686 \tab 442 \tab 193 \tab 47 \tab 12 \tab 19866 \cr
}
}

\seealso{
Other Core: 
\code{\link{chlddth}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{madeath}},
\code{\link{padeath}},
\code{\link{unrel1}},
\code{\link{unrel4}}
}
\concept{Core}
\keyword{variable}
