% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{decchurh}
\alias{decchurh}
\title{Imp of church teachings to make decision}
\source{
General Social Survey https://gss.norc.org
}
\description{
decchurh
}
\details{
Question Can you tell me how important is each of the following in helping you to make decisions about your life. Please indicate your answer by stating where you would place your feelings on a scale from 1 to 5 as shown on this hand Card.
C. The teachings of your church or synagogue
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[5]} not very important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1087/vshow}{GSS Data Explorer page for \code{decchurh}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab 2 \tab 3 \tab 4 \tab not very important \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 424 \tab 291 \tab 350 \tab 161 \tab 233 \tab 14 \tab 8 \tab 1481 \cr
   Total \tab 0 \tab 424 \tab 291 \tab 350 \tab 161 \tab 233 \tab 14 \tab 8 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{blkwhite}},
\code{\link{decbible}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{gochurch}},
\code{\link{permoral}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
