% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_200.R
\name{decgift}
\alias{decgift}
\title{Deciding on major charitable gifts.}
\source{
General Social Survey https://gss.norc.org
}
\description{
decgift
}
\details{
Question This question is about some of the financial decisions that you and your (husband/wife) may make. For each decision, please tell me how it was made the last time you faced this decision.
C. Deciding on major charitable gifts.
}
\section{Values}{

\itemize{
\item \code{1} husband alone
\item \code{2} husband after consulting with wife
\item \code{3} husband and wife together
\item \code{4} wife after consulting husband
\item \code{5} wife alone
\item \code{6} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1883/vshow}{GSS Data Explorer page for \code{decgift}}.
}

\seealso{
Other Markets: 
\code{\link{decbed}},
\code{\link{layoffs}}

Other Household: 
\code{\link{decbed}}

Other Purchases/Transactions: 
\code{\link{decbed}}
}
\concept{Household}
\concept{Markets}
\concept{Purchases/Transactions}
\keyword{variable}
