% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{deckids}
\alias{deckids}
\title{Who makes decision about how to bring up children}
\source{
General Social Survey https://gss.norc.org
}
\description{
deckids
}
\details{
Question Who usually makes/made the decisions about how to bring up your children?
}
\section{Values}{

\itemize{
\item \verb{[1]} mostly me
\item \verb{[2]} mostly my spouse
\item \verb{[3]} sometimes me or sometimes my spouse
\item \verb{[4]} we decide together
\item \verb{[5]} someone else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4439/vshow}{GSS Data Explorer page for \code{deckids}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab mostly me \tab mostly my spouse \tab sometimes me or sometimes my spouse \tab we decide together \tab someone else \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2058 \tab 81 \tab 44 \tab 102 \tab 305 \tab 6 \tab 145 \tab 24 \tab 2765 \cr
   2012 \tab 1044 \tab 233 \tab 65 \tab 96 \tab 481 \tab 3 \tab 42 \tab 10 \tab 1974 \cr
   Total \tab 3102 \tab 314 \tab 109 \tab 198 \tab 786 \tab 9 \tab 187 \tab 34 \tab 4739 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{spborn}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
