% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{deckids}
\alias{deckids}
\title{Who makes decision about how to bring up children}
\source{
General Social Survey https://gss.norc.org
}
\description{
deckids
}
\details{
Question Who usually makes/made the decisions about how to bring up your children?
}
\section{Values}{

\itemize{
\item \code{1} mostly me
\item \code{2} mostly my spouse
\item \code{3} sometimes me or sometimes my spouse
\item \code{4} we decide together
\item \code{5} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4439/vshow}{GSS Data Explorer page for \code{deckids}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{spborn}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\concept{Parents}
\keyword{variable}
