% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{decmoney}
\alias{decmoney}
\title{How able is x to decide managing own money}
\source{
General Social Survey https://gss.norc.org
}
\description{
decmoney
}
\details{
Question In your opinion, how able is NAME to make his/her own decisions about managing his/her own money -- very able, somewhat able, not very able or not able at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very able
\item \verb{[2]} somewhat able
\item \verb{[3]} not very able
\item \verb{[4]} not able at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2000/vshow}{GSS Data Explorer page for \code{decmoney}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very able \tab somewhat able \tab not very able \tab not able at all \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 281 \tab 384 \tab 429 \tab 292 \tab 43 \tab 15 \tab 2904 \cr
   2006 \tab 3073 \tab 338 \tab 484 \tab 383 \tab 163 \tab 54 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 256 \tab 387 \tab 338 \tab 138 \tab 26 \tab 28 \tab 2348 \cr
   Total \tab 5708 \tab 875 \tab 1255 \tab 1150 \tab 593 \tab 123 \tab 58 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
