% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{decmoney}
\alias{decmoney}
\title{How able is x to decide managing own money}
\source{
General Social Survey https://gss.norc.org
}
\description{
decmoney
}
\details{
Question In your opinion, how able is NAME to make his/her own decisions about managing his/her own money -- very able, somewhat able, not very able or not able at all?
}
\section{Values}{

\itemize{
\item \code{1} very able
\item \code{2} somewhat able
\item \code{3} not very able
\item \code{4} not able at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2000/vshow}{GSS Data Explorer page for \code{decmoney}}.
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
