% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{dectreat}
\alias{dectreat}
\title{How able is x to decide receive treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
dectreat
}
\details{
Question In your opinion, how able is NAME to make his/her own decisions about the treatment s/he should receive -- very able, somewhat able, not very able or not able at all?
}
\section{Values}{

\itemize{
\item \code{1} very able
\item \code{2} somewhat able
\item \code{3} not very able
\item \code{4} not able at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1999/vshow}{GSS Data Explorer page for \code{dectreat}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very able \tab somewhat able \tab not very able \tab not able at all \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 294 \tab 422 \tab 445 \tab 241 \tab 28 \tab 14 \tab 2904 \cr
   2006 \tab 3073 \tab 335 \tab 519 \tab 364 \tab 163 \tab 40 \tab 16 \tab 4510 \cr
   2018 \tab 1175 \tab 263 \tab 425 \tab 314 \tab 125 \tab 17 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 892 \tab 1366 \tab 1123 \tab 529 \tab 85 \tab 59 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
