% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{defpensn}
\alias{defpensn}
\title{R has defined benefit pension plan}
\source{
General Social Survey https://gss.norc.org
}
\description{
defpensn
}
\details{
Question B. Do you have a defined benefit pension plan at your company?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018 \cr
   B/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3562/vshow}{GSS Data Explorer page for \code{defpensn}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2014 \tab 1765 \tab 276 \tab 491 \tab 4 \tab 2 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 525 \tab 868 \tab 24 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 2570 \tab 316 \tab 591 \tab 52 \tab 6 \tab 9 \tab 3544 \cr
   Total \tab 5235 \tab 1117 \tab 1950 \tab 80 \tab 39 \tab 9 \tab 8430 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{trdestck}}
}
\concept{Shared Capitalism}
\keyword{variable}
