% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_84.R
\name{defspdfg}
\alias{defspdfg}
\title{Defense spending--govt position}
\source{
General Social Survey https://gss.norc.org
}
\description{
defspdfg
}
\details{
Question Some people believe that we should spend much less money for defense. Suppose these people are at one end of the scale at point number 1. Others feel that defense spending should be greatly increased. Suppose these people are at the other end, at point 7. And, of course, some other people have opinions somewhere in between at points 2, 3, 4, 5, or 6. ENTER CODE IN BOXES.
B. Where would you place what the federal government is doing at the present time?
}
\section{Values}{

\itemize{
\item \code{0} havent thought much
\item \code{1} greatly decrease
\item \code{7} greatly increase
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/760/vshow}{GSS Data Explorer page for \code{defspdfg}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab havent thought much \tab greatly decrease \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab greatly increase \tab don't know \tab no answer \tab Total \cr
   1980 \tab - \tab 84 \tab 63 \tab 128 \tab 317 \tab 347 \tab 207 \tab 123 \tab 92 \tab 99 \tab 8 \tab 1468 \cr
   Total \tab 0 \tab 84 \tab 63 \tab 128 \tab 317 \tab 347 \tab 207 \tab 123 \tab 92 \tab 99 \tab 8 \tab 1468 \cr
}
}

\seealso{
Other Core: 
\code{\link{defspdr}},
\code{\link{sibdeath}},
\code{\link{spdeath}},
\code{\link{trarel1}},
\code{\link{trarel5}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}
}
\concept{Core}
\keyword{variable}
