% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{defund}
\alias{defund}
\title{Move funds for police to social service}
\source{
General Social Survey https://gss.norc.org
}
\description{
defund
}
\details{
Question Favor or oppose reducing funding for police and instead funding social services
}
\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7574/vshow}{GSS Data Explorer page for \code{defund}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab favor \tab oppose \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1315 \tab 952 \tab 1392 \tab 7 \tab 332 \tab 34 \tab 4032 \cr
   Total \tab 1315 \tab 952 \tab 1392 \tab 7 \tab 332 \tab 34 \tab 4032 \cr
}
}

\keyword{variable}
