% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{defwrkev}
\alias{defwrkev}
\title{Ever work for military supplier?}
\source{
General Social Survey https://gss.norc.org
}
\description{
defwrkev
}
\details{
Question Have you ever worked for a company where a major part of their business was selling supplies or services to the armed forces?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/829/vshow}{GSS Data Explorer page for \code{defwrkev}}.
}

\seealso{
Other Core: 
\code{\link{guerilla}},
\code{\link{jobtrain}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()},
\code{\link{vetaid}}

Other Military: 
\code{\link{jobtrain}},
\code{\link{upgrade}()}
}
\concept{Core}
\concept{Military}
\keyword{variable}
