% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{defwrkev}
\alias{defwrkev}
\title{Ever work for military supplier?}
\source{
General Social Survey https://gss.norc.org
}
\description{
defwrkev
}
\details{
Question Have you ever worked for a company where a major part of their business was selling supplies or services to the armed forces?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/829/vshow}{GSS Data Explorer page for \code{defwrkev}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1984 \tab - \tab 181 \tab 1262 \tab 30 \tab 1473 \cr
   Total \tab 0 \tab 181 \tab 1262 \tab 30 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{guerilla}},
\code{\link{jobtrain}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()},
\code{\link{vetaid}}
}
\concept{Core}
\keyword{variable}
