% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_92.R
\name{defwrknw}
\alias{defwrknw}
\title{Currently work for military supplier?}
\source{
General Social Survey https://gss.norc.org
}
\description{
defwrknw
}
\details{
Question Have you ever worked for a company where a major part of their business was selling supplies or services to the armed forces?
A. Do you work for such a company now?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/830/vshow}{GSS Data Explorer page for \code{defwrknw}}.
}

\seealso{
Other Core: 
\code{\link{milwrkev}},
\code{\link{milwrknw}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{resdefwk}}
}
\concept{Core}
\keyword{variable}
