% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{dejavu}
\alias{dejavu}
\title{Thought one was someplace before}
\source{
General Social Survey https://gss.norc.org
}
\description{
dejavu
}
\details{
Question How often have you had any of the following experiences?  READ EACH ITEM AND CIRCLE ONE CODE FOR EACH.
A. Thought you were somewhere you had been before, but knew that it was impossible.
}
\section{Values}{

\itemize{
\item \code{1} never in my life
\item \code{2} once or twice
\item \code{3} several times
\item \code{4} often
\item \code{8} cannot answer
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/316/vshow}{GSS Data Explorer page for \code{dejavu}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never in my life \tab once or twice \tab several times \tab often \tab can't choose \tab no answer \tab Total \cr
   1984 \tab - \tab 482 \tab 369 \tab 449 \tab 139 \tab 15 \tab 19 \tab 1473 \cr
   1988 \tab - \tab 476 \tab 471 \tab 413 \tab 96 \tab 21 \tab 4 \tab 1481 \cr
   1989 \tab 531 \tab 359 \tab 293 \tab 271 \tab 67 \tab 13 \tab 3 \tab 1537 \cr
   Total \tab 531 \tab 1317 \tab 1133 \tab 1133 \tab 302 \tab 49 \tab 26 \tab 4491 \cr
}
}

\seealso{
Other Core: 
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
