% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{dem10fut}
\alias{dem10fut}
\title{How well will democracy work in america in ten yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
dem10fut
}
\details{
Question On the whole, on a scale of  0 to 10 where 0 is very poorly and 10 is very well.
C. And how about 10 years from now? How well do you think democracy will work in America then?
}
\section{Values}{

\itemize{
\item \verb{[0]} very poorly
\item \verb{[10]} very well
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4962/vshow}{GSS Data Explorer page for \code{dem10fut}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab very poorly \tab 1 \tab very well \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 72 \tab 30 \tab 150 \tab 53 \tab 90 \tab 109 \tab 181 \tab 131 \tab 197 \tab 250 \tab 106 \tab 94 \tab 9 \tab 2812 \cr
   2014 \tab 1274 \tab 66 \tab 45 \tab 74 \tab 74 \tab 81 \tab 91 \tab 180 \tab 108 \tab 168 \tab 117 \tab 59 \tab 151 \tab 50 \tab 2538 \cr
   Total \tab 2614 \tab 138 \tab 75 \tab 224 \tab 127 \tab 171 \tab 200 \tab 361 \tab 239 \tab 365 \tab 367 \tab 165 \tab 245 \tab 59 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
