% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{dem10pst}
\alias{dem10pst}
\title{How well did democracy work in america ten yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
dem10pst
}
\details{
Question On the whole, on a scale of  0 to 10 where 0 is very poorly and 10 is very well.
B. And what about 10 years ago? How well did democracy work in America then?
}
\section{Values}{

\itemize{
\item \code{0} very poorly
\item \code{10} very well
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4961/vshow}{GSS Data Explorer page for \code{dem10pst}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
