% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{dem10pst}
\alias{dem10pst}
\title{How well did democracy work in america ten yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
dem10pst
}
\details{
Question On the whole, on a scale of  0 to 10 where 0 is very poorly and 10 is very well.
B. And what about 10 years ago? How well did democracy work in America then?
}
\section{Values}{

\itemize{
\item \code{0} very poorly
\item \code{10} very well
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4961/vshow}{GSS Data Explorer page for \code{dem10pst}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab very poorly \tab 1 \tab very well \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 41 \tab 7 \tab 167 \tab 29 \tab 36 \tab 72 \tab 195 \tab 145 \tab 242 \tab 310 \tab 134 \tab 87 \tab 7 \tab 2812 \cr
   2014 \tab 1274 \tab 24 \tab 7 \tab 75 \tab 33 \tab 55 \tab 84 \tab 223 \tab 121 \tab 218 \tab 203 \tab 75 \tab 96 \tab 50 \tab 2538 \cr
   Total \tab 2614 \tab 65 \tab 14 \tab 242 \tab 62 \tab 91 \tab 156 \tab 418 \tab 266 \tab 460 \tab 513 \tab 209 \tab 183 \tab 57 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
