% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{demands}
\alias{demands}
\title{R feels that family make too many demands on r}
\source{
General Social Survey https://gss.norc.org
}
\description{
demands
}
\details{
Question Do you feel that your family, relatives and/or friends make too many demands on you?
}
\section{Values}{

\itemize{
\item \code{1} no, never
\item \code{2} yes, but seldom
\item \code{3} yes, sometimes
\item \code{4} yes, often
\item \code{5} yes, very often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4235/vshow}{GSS Data Explorer page for \code{demands}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab no, never \tab yes, but seldom \tab yes, sometimes \tab yes, often \tab yes, very often \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 504 \tab 282 \tab 269 \tab 54 \tab 27 \tab 3 \tab 10 \tab 2765 \cr
   2018 \tab 1175 \tab 696 \tab 228 \tab 177 \tab 44 \tab 24 \tab 2 \tab 2 \tab 2348 \cr
   Total \tab 2791 \tab 1200 \tab 510 \tab 446 \tab 98 \tab 51 \tab 5 \tab 12 \tab 5113 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{helpfrds}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
