% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{demrghts}
\alias{demrghts}
\title{Democratic rghts shouldn't be restricted by govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
demrghts
}
\details{
Question Here are some views regarding America's political system. Which of these statements is closer to your view?
}
\section{Values}{

\itemize{
\item \verb{[1]} shouldnt be restricted by govt
\item \verb{[2]} should be restricted if govt think necessary
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4963/vshow}{GSS Data Explorer page for \code{demrghts}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab shouldnt be restricted by govt \tab should be restricted if govt think necessary \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 991 \tab 373 \tab 85 \tab 23 \tab 2812 \cr
   Total \tab 1340 \tab 991 \tab 373 \tab 85 \tab 23 \tab 2812 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
