% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{demrghts}
\alias{demrghts}
\title{Democratic rghts shouldn't be restricted by govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
demrghts
}
\details{
Question Here are some views regarding America’s political system. Which of these statements is closer to your view?
}
\section{Values}{

\itemize{
\item \code{1} shouldnt be restricted by govt
\item \code{2} should be restricted if govt think necessary
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4963/vshow}{GSS Data Explorer page for \code{demrghts}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
