% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{demtoday}
\alias{demtoday}
\title{How well democracy work in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
demtoday
}
\details{
Question On the whole, on a scale of  0 to 10 where 0 is very poorly and 10 is very well.
A. How well does democracy work in America today?
}
\section{Values}{

\itemize{
\item \code{0} very poorly
\item \code{10} very well
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4960/vshow}{GSS Data Explorer page for \code{demtoday}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab very poorly \tab 1 \tab very well \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 50 \tab 5 \tab 140 \tab 23 \tab 58 \tab 67 \tab 227 \tab 170 \tab 277 \tab 316 \tab 108 \tab 25 \tab 6 \tab 2812 \cr
   2014 \tab 1274 \tab 32 \tab 13 \tab 55 \tab 44 \tab 77 \tab 83 \tab 249 \tab 177 \tab 235 \tab 164 \tab 45 \tab 64 \tab 26 \tab 2538 \cr
   Total \tab 2614 \tab 82 \tab 18 \tab 195 \tab 67 \tab 135 \tab 150 \tab 476 \tab 347 \tab 512 \tab 480 \tab 153 \tab 89 \tab 32 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
