% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{demworks}
\alias{demworks}
\title{How well the system of democracy in america works?}
\source{
General Social Survey https://gss.norc.org
}
\description{
demworks
}
\details{
Question All in all, how well or badly do you think the system of democracy in America works these days? PLEASE CHECK OnE BOX ONLY.
}
\section{Values}{

\itemize{
\item \verb{[1]} it works well and needs no changes
\item \verb{[2]} it works well but needs some changes
\item \verb{[3]} needs a lot of changes
\item \verb{[4]} needs to be completely changed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4102/vshow}{GSS Data Explorer page for \code{demworks}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab it works well and needs no changes \tab it works well but needs some changes \tab needs a lot of changes \tab needs to be completely changed \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 57 \tab 832 \tab 299 \tab 58 \tab 62 \tab 24 \tab 2904 \cr
   Total \tab 1572 \tab 57 \tab 832 \tab 299 \tab 58 \tab 62 \tab 24 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
