% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{demworks}
\alias{demworks}
\title{How well the system of democracy in america works?}
\source{
General Social Survey https://gss.norc.org
}
\description{
demworks
}
\details{
Question All in all, how well or badly do you think the system of democracy in America works these days? PLEASE CHECK OnE BOX ONLY.
}
\section{Values}{

\itemize{
\item \code{1} it works well and needs no changes
\item \code{2} it works well but needs some changes
\item \code{3} needs a lot of changes
\item \code{4} needs to be completely changed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4102/vshow}{GSS Data Explorer page for \code{demworks}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}

Other Federal Government: 
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxspend}}

Other Political: 
\code{\link{poleff17}}
}
\concept{Federal Government}
\concept{ISSP Role of Gov't}
\concept{Political}
\concept{United States}
\keyword{variable}
