% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{denom16}
\alias{denom16}
\title{Denomination in which r was raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
denom16
}
\details{
Question In what religion were you raised?  A. What specific denomination is that, if any?
As of 2024, DENOM16 has been recoded to only account for seven broad denomination categories. This has been retroactively applied to all previous years of DENOM16. The original DENOM16 containing data 1973 to 2022 has been renamed DENOM16_7322.
}
\section{Values}{

\itemize{
\item \code{10} american baptist association
\item \code{11} american baptist churches in the u.s.a.
\item \code{12} national baptist convention of america
\item \code{13} national baptist convention u.s.a., inc.
\item \code{14} southern baptist convention
\item \code{15} other baptist churches
\item \code{18} baptist, don't know which
\item \code{20} african methodist episcopal church
\item \code{21} african methodist episcopal zion church
\item \code{22} united methodist church
\item \code{23} other methodist churches
\item \code{28} methodist, don't know which
\item \code{30} american lutheran church
\item \code{31} lutheran church in america
\item \code{32} lutheran church - missouri synod
\item \code{33} wisconsin evangelical lutheran synod
\item \code{34} other lutheran churches
\item \code{35} evangelical lutheran
\item \code{38} lutheran, don't know which
\item \code{40} presbyterian church in the u.s.
\item \code{41} united presbyterian church in the u.s.a.
\item \code{42} other presbyterian churches
\item \code{43} presbyterian, merged
\item \code{48} presbyterian, don't know which
\item \code{50} episcopal church
\item \code{60} other, specify
\item \code{70} no denomination or non-denominational church
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/342/vshow}{GSS Data Explorer page for \code{denom16}}.
}

\seealso{
Other Core: 
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}

Other Religion: 
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
