% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{denom_next}
\alias{denom_next}
\title{Denomination - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
denom_next
}
\details{
Question What specific denomination is that, if any?
}
\section{Values}{

\itemize{
\item \code{10} american baptist association
\item \code{11} american baptist churches in the u.s.a.
\item \code{12} national baptist convention of america
\item \code{13} national baptist convention u.s.a., inc.
\item \code{14} southern baptist convention
\item \code{15} other baptist churches
\item \code{18} baptist, don't know which
\item \code{20} african methodist episcopal church
\item \code{21} african methodist episcopal zion church
\item \code{22} united methodist church
\item \code{23} other methodist churches
\item \code{28} methodist, don't know which
\item \code{30} american lutheran church
\item \code{31} lutheran church in america
\item \code{32} lutheran church - missouri synod
\item \code{33} wisconsin evangelical lutheran synod
\item \code{34} other lutheran churches
\item \code{35} evangelical lutheran
\item \code{38} lutheran, don't know which
\item \code{40} presbyterian church in the u.s.
\item \code{41} united presbyterian church in the u.s.a.
\item \code{42} other presbyterian churches
\item \code{43} presbyterian, merged
\item \code{48} presbyterian, don't know which
\item \code{50} episcopal church
\item \code{60} other, specify
\item \code{70} no denomination or non-denominational church
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8002/vshow}{GSS Data Explorer page for \code{denom_next}}.
}

\keyword{variable}
