% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{denomsch}
\alias{denomsch}
\title{Denomination of school children attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
denomsch
}
\details{
Question Do/did all of your children attend public school for all of their elementary and high school education?
A. Which types of school do/did your children attend?
3a. What denomination or church was/is this school affiliated with?
}
\section{Values}{

\itemize{
\item \code{1} baptist
\item \code{2} christian reform
\item \code{3} christian science
\item \code{4} episcopal
\item \code{5} inter denominational
\item \code{6} lutheran
\item \code{7} missouri synod lutheran
\item \code{8} non-denominational
\item \code{9} southern baptist
\item \code{10} seventh day adventist
\item \code{11} valley christian
\item \code{12} congregational
\item \code{13} mormon
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1151/vshow}{GSS Data Explorer page for \code{denomsch}}.
}

\seealso{
Other Religion: 
\code{\link{chrissch}},
\code{\link{implives}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}}

Other Education: 
\code{\link{chrissch}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}}
}
\concept{Education}
\concept{Religion}
\keyword{variable}
