% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{densp}
\alias{densp}
\title{If protestant, r's spouse's denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
densp
}
\details{
Question What is your husband's/wife's religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
A. If Protestant, what specific denomination was that?
}
\section{Values}{

\itemize{
\item \code{10000} amer baptist assoc
\item \code{11000} amer baptist ch in usa
\item \code{12000} natl baptist conv. of america
\item \code{13000} natl baptist conv., usa, inc
\item \code{14000} southern baptist conv
\item \code{15000} other baptist chs
\item \code{18000} baptist, dk which
\item \code{20000} african meth epis ch
\item \code{21000} ame zion church
\item \code{22000} united meth church
\item \code{23000} other meth ch
\item \code{28000} methodist, dk which
\item \code{30000} amer lutheran ch
\item \code{31000} luth ch in amer
\item \code{32000} luth ch--mo synod
\item \code{33000} wisc evan luth synod
\item \code{34000} other luth chs
\item \code{35000} evan lutheran
\item \code{38000} lutheran, dk which
\item \code{40000} presbyt ch in usa
\item \code{41000} united presbyt ch in usa
\item \code{42000} other presbyt chs
\item \code{43000} presbyt ch usa
\item \code{48000} presbyterian, dk which
\item \code{50000} episcopal church
\item \code{60002} evangeligal congregational
\item \code{60003} ind. bible fellowship
\item \code{60006} new testament christian
\item \code{60008} moravian
\item \code{60009} christ. and mission. alliance
\item \code{60010} adventist
\item \code{60012} assembly of god
\item \code{60013} free methodist
\item \code{60014} apostolic
\item \code{60015} african methodist
\item \code{60016} free will baptist
\item \code{60018} holiness, nazarene
\item \code{60020} brethren church
\item \code{60023} united brethren, breth. in christ
\item \code{60024} independent
\item \code{60031} christian; central christian
\item \code{60032} christian reform
\item \code{60033} christian scientist
\item \code{60035} church of christ
\item \code{60036} churches of god -exc. christ & holiness
\item \code{60037} church of god in christ
\item \code{60038} holiness
\item \code{60040} congregationalist, 1st cong
\item \code{60041} community church
\item \code{60042} covenant
\item \code{60043} dutch reform
\item \code{60044} disciples of christ
\item \code{60045} evangelical, evangelist
\item \code{60046} evangelical reformed
\item \code{60047} evangelist free church
\item \code{60048} first church
\item \code{60049} 1st christian disc of christ
\item \code{60050} 1st reformed
\item \code{60051} 1st christian
\item \code{60052} full gospel
\item \code{60056} holiness; ch of holiness
\item \code{60057} pilgrim holiness
\item \code{60058} jehovah witnesses
\item \code{60059} lds
\item \code{60060} lds--mormon
\item \code{60061} lds--reorganized
\item \code{60062} lds--jesus christ; ch of jesus lds
\item \code{60063} mennonite
\item \code{60064} mormon
\item \code{60065} nazarene
\item \code{60066} pentecostal assembly of god
\item \code{60068} pentecostal
\item \code{60070} quaker
\item \code{60071} reformed
\item \code{60076} salvation army
\item \code{60077} 7th day adventist
\item \code{60078} sanctified, sanctification
\item \code{60080} unitarian, universalist
\item \code{60081} united church of christ
\item \code{60082} united church, unity church
\item \code{60083} wesleyan
\item \code{60084} wesleyan methodist--pilgrim
\item \code{60092} mission covenant
\item \code{60093} missionary baptist
\item \code{60097} other fundamentalist
\item \code{60098} federated church
\item \code{60102} charismatic christian
\item \code{60112} evangeligal methodist
\item \code{60113} worldwide church of god
\item \code{60117} missionary
\item \code{60119} united church of canada
\item \code{60120} evangelical united brethren
\item \code{60133} primitive baptist
\item \code{60138} apostolic church
\item \code{60141} new birth christian
\item \code{60149} polish catholic
\item \code{60150} zwinglian
\item \code{60156} living church of god
\item \code{60181} christian born again
\item \code{60184} church of god israelite
\item \code{70000} no denom. given or non-denom ch
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4672/vshow}{GSS Data Explorer page for \code{densp}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{afterlif}},
\code{\link{ancestrs}},
\code{\link{antirel}},
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{befair}},
\code{\link{bible1}},
\code{\link{cantrust}},
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{comfort}},
\code{\link{conbiz}},
\code{\link{conchurh}},
\code{\link{concong}},
\code{\link{concourt}},
\code{\link{congovt}},
\code{\link{conschls}},
\code{\link{deathpen}},
\code{\link{denkid}},
\code{\link{devil}},
\code{\link{egomeans}},
\code{\link{fatalism}},
\code{\link{feelrel}},
\code{\link{geomobil}},
\code{\link{godchnge}},
\code{\link{godmeans}},
\code{\link{godright}},
\code{\link{govcheat}},
\code{\link{hapunhap}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{homosex1}},
\code{\link{madenkid}},
\code{\link{makefrnd}},
\code{\link{marelkid}},
\code{\link{miracles}},
\code{\link{mywaygod}},
\code{\link{nihilism}},
\code{\link{nirvana}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{paxhappy}},
\code{\link{perright}},
\code{\link{perscrfc}},
\code{\link{polsgod}},
\code{\link{prayfreq}},
\code{\link{predeter}},
\code{\link{premars1}},
\code{\link{reincar}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{relcand}},
\code{\link{relexper}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relgrpeq}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{religkid}},
\code{\link{religpub}},
\code{\link{religsp}},
\code{\link{relmarry}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{reltruth}},
\code{\link{rightpeo}},
\code{\link{rspctrel}},
\code{\link{schlpray}},
\code{\link{socright}},
\code{\link{stiffpun}},
\code{\link{taxcheat}},
\code{\link{theism}},
\code{\link{trustsci}},
\code{\link{vistholy}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}},
\code{\link{xmarsex1}}
}
\concept{ISSP Religion}
\keyword{variable}
