% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{denyrais}
\alias{denyrais}
\title{Denied raise without reason at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
denyrais
}
\details{
Question I have been denied a raise or promotion without being given a valid reason.
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3021/vshow}{GSS Data Explorer page for \code{denyrais}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2012 \tab 795 \tab 32 \tab 101 \tab 105 \tab 897 \tab 13 \tab 31 \tab 1974 \cr
   Total \tab 795 \tab 32 \tab 101 \tab 105 \tab 897 \tab 13 \tab 31 \tab 1974 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{eharaswk}},
\code{\link{ignorwk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{wkbhvrs}},
\code{\link{wkrspns}}
}
\concept{Workplace Conflict}
\keyword{variable}
