% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{devil}
\alias{devil}
\title{Belief in the devil}
\source{
General Social Survey https://gss.norc.org
}
\description{
devil
}
\details{
Question Do you believe in. . .
B. The Devil
}
\section{Values}{

\itemize{
\item \code{1} yes, definitely
\item \code{2} yes, probably
\item \code{3} no, probably not
\item \code{4} no, definitely not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4652/vshow}{GSS Data Explorer page for \code{devil}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}

Other Religion: 
\code{\link{bible1}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
