% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{difrace1}
\alias{difrace1}
\title{Why observation disagrees self id}
\source{
General Social Survey https://gss.norc.org
}
\description{
difrace1
}
\details{
Question A. WHY DO YOU THINK YOUR RACIAL CLASSIFICATION OF THE RESPONDENT IN "A" DIFFERED FROM THE RESPONDENT'S SELF-CLASSIFICATION IN "C"?
\enumerate{
\item First reason
}
}
\section{Values}{

\itemize{
\item \code{1} r not serious
\item \code{2} hispanic chosen over white
\item \code{3} hispanic chosen over black
\item \code{4} physically doesn't match race
\item \code{5} racially mixed
\item \code{6} r rejects racial categories
\item \code{7} r rejects 'black' label
\item \code{8} is married to member of another race
\item \code{9} white or black chosen over hispanic
\item \code{10} feelings, not biology matters
\item \code{11} identifies with other than 'real' race
\item \code{12} other hispanic complications
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5280/vshow}{GSS Data Explorer page for \code{difrace1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab r not serious \tab hispanic chosen over white \tab hispanic chosen over black \tab physically doesn't match race \tab racially mixed \tab r rejects racial categories \tab r rejects 'black' label \tab is married to member of another race \tab white or black chosen over hispanic \tab Total \cr
   1996 \tab 2865 \tab 1 \tab 15 \tab 3 \tab 4 \tab 5 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2904 \cr
   Total \tab 2865 \tab 1 \tab 15 \tab 3 \tab 4 \tab 5 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\keyword{variable}
