% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{difrace2}
\alias{difrace2}
\title{Why observation disagrees self id}
\source{
General Social Survey https://gss.norc.org
}
\description{
difrace2
}
\details{
Question A. WHY DO YOU THINK YOUR RACIAL CLASSIFICATION OF THE RESPONDENT IN "A" DIFFERED FROM THE RESPONDENT'S SELF-CLASSIFICATION IN "C"?
2. Second reason
}
\section{Values}{

\itemize{
\item \code{1} r not serious
\item \code{2} hispanic chosen over white
\item \code{3} hispanic chosen over black
\item \code{4} physically doesn't match race
\item \code{5} racially mixed
\item \code{6} r rejects racial categories
\item \code{7} r rejects 'black' label
\item \code{8} is married to member of another race
\item \code{9} white or black chosen over hispanic
\item \code{10} feelings, not biology matters
\item \code{11} identifies with other than 'real' race
\item \code{12} other hispanic complications
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5281/vshow}{GSS Data Explorer page for \code{difrace2}}.
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{fucitzn}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}}
}
\concept{Ethnicity}
\keyword{variable}
