% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{difstand}
\alias{difstand}
\title{Some ppl hold standard in workplace that others don?}
\source{
General Social Survey https://gss.norc.org
}
\description{
difstand
}
\details{
Question A. Some people are held to standards in my workplace that others are not.
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2998/vshow}{GSS Data Explorer page for \code{difstand}}.
}

\seealso{
Other Workplace Conflict: 
\code{\link{lackinfo}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{putdown}},
\code{\link{rptprobs}}
}
\concept{Workplace Conflict}
\keyword{variable}
