% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{dipged}
\alias{dipged}
\title{Diploma, ged, or other}
\source{
General Social Survey https://gss.norc.org
}
\description{
dipged
}
\details{
Question Was that a diploma awarded by your high school at graduation at the end of 12th grade, a GED awarded after you took a test, or something else?
As of 2024, DIPGED has been recoded to only account for three categories. This has been retroactively applied to all previous years of DIPGED. The original DIPGED containing data 2012 to 2022 has been renamed DIPGED_1222.
}
\section{Values}{

\itemize{
\item \code{1} high school diploma
\item \code{2} ged
\item \code{3} other
\item \code{5} high school diploma after post high school class
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/65/vshow}{GSS Data Explorer page for \code{dipged}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab high school diploma \tab ged \tab other \tab high school diploma after post high school class \tab no answer \tab don't know \tab skipped on web \tab 7 \tab Total \cr
   2012 \tab 288 \tab 1543 \tab 133 \tab 2 \tab 6 \tab 2 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 330 \tab 1990 \tab 196 \tab 15 \tab - \tab 6 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 328 \tab 2237 \tab 245 \tab 41 \tab - \tab 15 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 262 \tab 1891 \tab 183 \tab 3 \tab 3 \tab 6 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 269 \tab 3396 \tab 306 \tab 13 \tab 9 \tab 24 \tab 8 \tab 7 \tab - \tab 4032 \cr
   2022 \tab 359 \tab 2858 \tab 276 \tab 29 \tab 3 \tab 13 \tab 1 \tab 3 \tab 2 \tab 3544 \cr
   Total \tab 1836 \tab 13915 \tab 1339 \tab 103 \tab 21 \tab 66 \tab 11 \tab 10 \tab 2 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{codipged}},
\code{\link{cosector}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{major2}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\keyword{variable}
