% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_407.R
\name{disabld4}
\alias{disabld4}
\title{R has other physical disability}
\source{
General Social Survey https://gss.norc.org
}
\description{
disabld4
}
\details{
Question Do you have any other physical disability?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C/D \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3715/vshow}{GSS Data Explorer page for \code{disabld4}}.
}

\seealso{
Other Disability: 
\code{\link{disabld1}},
\code{\link{disabld2}},
\code{\link{disabld3}},
\code{\link{disabld5}},
\code{\link{disabld6}},
\code{\link{disabld7}}
}
\concept{Disability}
\keyword{variable}
