% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_651.R
\name{disblty}
\alias{disblty}
\title{Does r have disability}
\source{
General Social Survey https://gss.norc.org
}
\description{
disblty
}
\details{
Question Do you have a long-standing illness, a chronic condition, or a disability?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7838/vshow}{GSS Data Explorer page for \code{disblty}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2382 \tab 433 \tab 716 \tab 6 \tab 2 \tab 5 \tab 3544 \cr
   Total \tab 2382 \tab 433 \tab 716 \tab 6 \tab 2 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
