% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{discaffw}
\alias{discaffw}
\title{A woman won't get a job or promotion}
\source{
General Social Survey https://gss.norc.org
}
\description{
discaffw
}
\details{
Question B. What do you think the chances are these days that a woman won't get a job or promotion while an equally or less qualified man gets one instead? Is this very likely, somewhat likely, somewhat unlikely, or very unlikely these days?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1675/vshow}{GSS Data Explorer page for \code{discaffw}}.
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{chldben}},
\code{\link{discaffm}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{parleave}},
\code{\link{womenben}}
}
\concept{Gender}
\keyword{variable}
