% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{discaffwv}
\alias{discaffwv}
\title{A woman won't get a job or promotion (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
discaffwv
}
\details{
Question What do you think the chances are these days that a woman won't get a job or promotion while an equally or less qualified man gets one instead. Is this very likely, somewhat likely, somewhat unlikely, or very unlikely these days
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} very unlikely
\item \verb{[5]} no opinion (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7587/vshow}{GSS Data Explorer page for \code{discaffwv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab very unlikely \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2658 \tab 340 \tab 587 \tab 256 \tab 127 \tab 62 \tab 2 \tab 4032 \cr
   2022 \tab 2990 \tab 121 \tab 250 \tab 97 \tab 52 \tab 33 \tab 1 \tab 3544 \cr
   Total \tab 5648 \tab 461 \tab 837 \tab 353 \tab 179 \tab 95 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
