% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{discaffwv}
\alias{discaffwv}
\title{A woman won't get a job or promotion (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
discaffwv
}
\details{
Question What do you think the chances are these days that a woman won't get a job or promotion while an equally or less qualified man gets one instead. Is this very likely, somewhat likely, somewhat unlikely, or very unlikely these days
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} very unlikely
\item \code{5} no opinion (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7587/vshow}{GSS Data Explorer page for \code{discaffwv}}.
}

\keyword{variable}
