% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{discaffy}
\alias{discaffy}
\title{Chance blacks promoted ahead of whites}
\source{
General Social Survey https://gss.norc.org
}
\description{
discaffy
}
\details{
Question B. What do you think the chances are these days that a white person won't get a job or promotion while a less qualified black person gets one instead?  Is this very likely, somewhat likely, or not very likely to happen these days.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1624/vshow}{GSS Data Explorer page for \code{discaffy}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{ethhist}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
