% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_328.R
\name{discipln}
\alias{discipln}
\title{Situation caused by lacking discipline in the}
\source{
General Social Survey https://gss.norc.org
}
\description{
discipln
}
\details{
Question G. In your opinion, how likely is it that (John/Mary)'s situation MIGHT be caused by lacking discipline in the home?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2927/vshow}{GSS Data Explorer page for \code{discipln}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 219 \tab 535 \tab 317 \tab 279 \tab 33 \tab 10 \tab 2765 \cr
   Total \tab 1372 \tab 219 \tab 535 \tab 317 \tab 279 \tab 33 \tab 10 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{allergic}},
\code{\link{bettrown}},
\code{\link{geneprob}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{nextdoor}},
\code{\link{raised}},
\code{\link{violtv}}
}
\concept{Mental Health II}
\keyword{variable}
