% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_134.R
\name{discwhy6}
\alias{discwhy6}
\title{Other sources}
\source{
General Social Survey https://gss.norc.org
}
\description{
discwhy6
}
\details{
Question What do you think the chances are these days that a white person won't get a job or promotion while an equally or less qualified black person gets one instead?  Is this very likely, somewhat likely, or not very likely to happen these days?
A. Do you feel this way because of something that happened to you personally, because it happened to a relative, family member or close friend, because you saw it occurring at work, or because you have heard about it from the media or other sources? (CHECK ALL THAT APPLY)
6. Other (SPECIFY)
}
\section{Values}{

\itemize{
\item \code{1} general,no source
\item \code{2} all of the above
\item \code{3} the law
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1257/vshow}{GSS Data Explorer page for \code{discwhy6}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{affact}},
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}}

Other Multiculturalism: 
\code{\link{discaff}},
\code{\link{discwhy1}},
\code{\link{discwhy2}},
\code{\link{discwhy3}},
\code{\link{discwhy4}},
\code{\link{discwhy5}}
}
\concept{Intergroup Relations}
\concept{Multiculturalism}
\keyword{variable}
