% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{divbest}
\alias{divbest}
\title{Divorce as best solution to marital problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
divbest
}
\details{
Question Do you agree or disagree…
Divorce is usually the best solution when a couple can''t seem to work out their marriage problems.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4408/vshow}{GSS Data Explorer page for \code{divbest}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 126 \tab 548 \tab 272 \tab 334 \tab 112 \tab 33 \tab 22 \tab 2992 \cr
   2002 \tab 1594 \tab 156 \tab 338 \tab 233 \tab 302 \tab 120 \tab 21 \tab 1 \tab 2765 \cr
   2012 \tab 672 \tab 91 \tab 561 \tab 187 \tab 336 \tab 90 \tab 29 \tab 8 \tab 1974 \cr
   Total \tab 3811 \tab 373 \tab 1447 \tab 692 \tab 972 \tab 322 \tab 83 \tab 31 \tab 7731 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabfst}},
\code{\link{cohabok}},
\code{\link{divifkd1}},
\code{\link{divnokd1}},
\code{\link{fewrksup}},
\code{\link{hubbywk1}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{singlpar}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
