% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{divifkid}
\alias{divifkid}
\title{How easy should divorce be if kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
divifkid
}
\details{
Question And what about couples with young children? How easy or difficult should the law make it for them to get a divorce?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} fairly easy
\item \code{3} neither easy nor difficult
\item \code{4} fairly difficult
\item \code{5} very difficult
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4388/vshow}{GSS Data Explorer page for \code{divifkid}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
