% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{divkids}
\alias{divkids}
\title{Kids better if bad marrge ended by div}
\source{
General Social Survey https://gss.norc.org
}
\description{
divkids
}
\details{
Question When a marriage is troubled and unhappy, do you think it is generally better for the children if the couple stays together or gets divorced?
}
\section{Values}{

\itemize{
\item \code{1} much better to div
\item \code{2} better to divorce
\item \code{3} worse to divorce
\item \code{4} much worse to div
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4389/vshow}{GSS Data Explorer page for \code{divkids}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab much better to div \tab better to divorce \tab worse to divorce \tab much worse to div \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 238 \tab 536 \tab 194 \tab 105 \tab 315 \tab 26 \tab 1481 \cr
   Total \tab 67 \tab 238 \tab 536 \tab 194 \tab 105 \tab 315 \tab 26 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
