% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{divlawnv}
\alias{divlawnv}
\title{Divorce laws (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
divlawnv
}
\details{
Question Should divorce in this country be easier or more difficult to obtain than it is now?
}
\section{Values}{

\itemize{
\item \code{1} easier
\item \code{2} more difficult
\item \code{3} stay as is (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7610/vshow}{GSS Data Explorer page for \code{divlawnv}}.
}

\keyword{variable}
