% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{divorce5}
\alias{divorce5}
\title{Divorced in last 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
divorce5
}
\details{
Question Divorces during the last five years.
}
\section{Values}{

\itemize{
\item \code{0} none
\item \code{1} prior 4 years
\item \code{2} last year
\item \code{3} both
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/737/vshow}{GSS Data Explorer page for \code{divorce5}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab none \tab prior 4 years \tab last year \tab both \tab no answer \tab Total \cr
   1978 \tab - \tab 1353 \tab 116 \tab 40 \tab 9 \tab 14 \tab 1532 \cr
   1980 \tab - \tab 1318 \tab 88 \tab 27 \tab 9 \tab 26 \tab 1468 \cr
   1983 \tab - \tab 1431 \tab 112 \tab 32 \tab 17 \tab 7 \tab 1599 \cr
   1984 \tab - \tab 1319 \tab 86 \tab 35 \tab 8 \tab 25 \tab 1473 \cr
   1986 \tab - \tab 1310 \tab 88 \tab 28 \tab 8 \tab 36 \tab 1470 \cr
   1987 \tab - \tab 1592 \tab 98 \tab 33 \tab 8 \tab 88 \tab 1819 \cr
   1988 \tab 484 \tab 882 \tab 55 \tab 19 \tab 13 \tab 28 \tab 1481 \cr
   1989 \tab 502 \tab 912 \tab 53 \tab 18 \tab 10 \tab 42 \tab 1537 \cr
   1990 \tab 473 \tab 790 \tab 56 \tab 19 \tab 3 \tab 31 \tab 1372 \cr
   1991 \tab 500 \tab 897 \tab 44 \tab 24 \tab 7 \tab 45 \tab 1517 \cr
   1993 \tab 549 \tab 930 \tab 67 \tab 25 \tab 11 \tab 24 \tab 1606 \cr
   1994 \tab 2481 \tab 454 \tab 25 \tab 12 \tab 8 \tab 12 \tab 2992 \cr
   Total \tab 4989 \tab 13188 \tab 888 \tab 312 \tab 111 \tab 378 \tab 19866 \cr
}
}

\seealso{
Other Core: 
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{unemp5}}
}
\concept{Core}
\keyword{variable}
