% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{divrel1}
\alias{divrel1}
\title{Relatives divorced in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
divrel1
}
\details{
Question A. Number of relatives divorced last year.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/738/vshow}{GSS Data Explorer page for \code{divrel1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab Total \cr
   1978 \tab - \tab 1337 \tab 171 \tab 20 \tab 3 \tab 1 \tab 1532 \cr
   1980 \tab - \tab 1281 \tab 172 \tab 15 \tab - \tab - \tab 1468 \cr
   1983 \tab - \tab 1373 \tab 183 \tab 40 \tab 3 \tab - \tab 1599 \cr
   1984 \tab - \tab 1300 \tab 103 \tab 66 \tab 2 \tab 2 \tab 1473 \cr
   Total \tab 0 \tab 5291 \tab 629 \tab 141 \tab 8 \tab 3 \tab 6072 \cr
}
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}},
\code{\link{unemp5}}
}
\concept{Core}
\keyword{variable}
