% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{divwife}
\alias{divwife}
\title{Wife better if bad marrge ended by div}
\source{
General Social Survey https://gss.norc.org
}
\description{
divwife
}
\details{
Question And when a marriage is troubled and unhappy, is it generally better for the wife if the couple stays together or gets divorced?
}
\section{Values}{

\itemize{
\item \verb{[1]} much better to div
\item \verb{[2]} better to divorce
\item \verb{[3]} worse to divorce
\item \verb{[4]} much worse to div
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4390/vshow}{GSS Data Explorer page for \code{divwife}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab much better to div \tab better to divorce \tab worse to divorce \tab much worse to div \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 232 \tab 558 \tab 185 \tab 83 \tab 326 \tab 30 \tab 1481 \cr
   Total \tab 67 \tab 232 \tab 558 \tab 185 \tab 83 \tab 326 \tab 30 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
