% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_244.R
\name{doc1}
\alias{doc1}
\title{Doctors aren't as thorough as should be}
\source{
General Social Survey https://gss.norc.org
}
\description{
doc1
}
\details{
Question As you read each of the following statements, please think about the medical care you are now receiving. If you have not received any medical care recently, circle the answer based on what you would expect if you had to seek care today. Even if you are not entirely certain about your answers, we want to remind you that your best guess is important for each statement.
A. Doctors aren't as thorough as they should be.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} uncertain
\item \verb{[4]} disagree
\item \verb{[5]} strong disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2200/vshow}{GSS Data Explorer page for \code{doc1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab uncertain \tab disagree \tab strong disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 128 \tab 571 \tab 191 \tab 425 \tab 48 \tab 21 \tab 3 \tab 2832 \cr
   Total \tab 1445 \tab 128 \tab 571 \tab 191 \tab 425 \tab 48 \tab 21 \tab 3 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{mntldeny}},
\code{\link{mntldif}}
}
\concept{Mental Health}
\keyword{variable}
