% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_649.R
\name{docalt}
\alias{docalt}
\title{How often visit alt health care practitioner}
\source{
General Social Survey https://gss.norc.org
}
\description{
docalt
}
\details{
Question (During the past 12 months, how often did you visit or were visited by... ) an alternative health care practitioner?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} seldom
\item \verb{[3]} sometimes
\item \verb{[4]} often
\item \verb{[5]} very often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7826/vshow}{GSS Data Explorer page for \code{docalt}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab seldom \tab sometimes \tab often \tab very often \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2381 \tab 815 \tab 134 \tab 115 \tab 52 \tab 12 \tab 25 \tab 4 \tab 6 \tab 3544 \cr
   Total \tab 2381 \tab 815 \tab 134 \tab 115 \tab 52 \tab 12 \tab 25 \tab 4 \tab 6 \tab 3544 \cr
}
}

\keyword{variable}
