% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{docaskme}
\alias{docaskme}
\title{R prefers doc ask opinion}
\source{
General Social Survey https://gss.norc.org
}
\description{
docaskme
}
\details{
Question I will read you some statements of beliefs people have.  Please look at the card and decide which answer best applies to you.
A. I prefer that my doctor offers me choices and asks my opinion.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} moderately agree
\item \code{3} slightly agree
\item \code{4} slightly disagree
\item \code{5} moderately disagree
\item \code{6} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2740/vshow}{GSS Data Explorer page for \code{docaskme}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab moderately agree \tab slightly agree \tab slightly disagree \tab moderately disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 1986 \tab 521 \tab 145 \tab 43 \tab 31 \tab 19 \tab 15 \tab 5 \tab 2765 \cr
   Total \tab 0 \tab 1986 \tab 521 \tab 145 \tab 43 \tab 31 \tab 19 \tab 15 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{docvig}},
\code{\link{regdoc}},
\code{\link{regdocyr}}
}
\concept{Doctors & Patients}
\keyword{variable}
