% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{doccosts}
\alias{doccosts}
\title{How r trust doc put health above costs}
\source{
General Social Survey https://gss.norc.org
}
\description{
doccosts
}
\details{
Question How much would you trust the doctor to put your health above costs? Would you say...
}
\section{Values}{

\itemize{
\item \code{1} completely
\item \code{2} mostly
\item \code{3} somewhat
\item \code{4} a little
\item \code{5} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2756/vshow}{GSS Data Explorer page for \code{doccosts}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{insrchng}},
\code{\link{secopin}}

Other Health: 
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{insrchng}},
\code{\link{secopin}}
}
\concept{Doctors & Patients}
\concept{Health}
\keyword{variable}
