% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{doccosts}
\alias{doccosts}
\title{How r trust doc put health above costs}
\source{
General Social Survey https://gss.norc.org
}
\description{
doccosts
}
\details{
Question How much would you trust the doctor to put your health above costs? Would you say...
}
\section{Values}{

\itemize{
\item \code{1} completely
\item \code{2} mostly
\item \code{3} somewhat
\item \code{4} a little
\item \code{5} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2756/vshow}{GSS Data Explorer page for \code{doccosts}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab completely \tab mostly \tab somewhat \tab a little \tab not at all \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 362 \tab 647 \tab 748 \tab 377 \tab 585 \tab 41 \tab 5 \tab 2765 \cr
   Total \tab 0 \tab 362 \tab 647 \tab 748 \tab 377 \tab 585 \tab 41 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{insrchng}},
\code{\link{secopin}}
}
\concept{Doctors & Patients}
\keyword{variable}
