% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{docrely}
\alias{docrely}
\title{R relies on doctors knowledge}
\source{
General Social Survey https://gss.norc.org
}
\description{
docrely
}
\details{
Question I will read you some statements of beliefs people have.  Please look at the card and decide which answer best applies to you.
C. I prefer to rely on my doctor's knowledge and not try to find out about my condition on my own.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} moderately agree
\item \code{3} slightly agree
\item \code{4} slightly disagree
\item \code{5} moderately disagree
\item \code{6} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2742/vshow}{GSS Data Explorer page for \code{docrely}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docvig}},
\code{\link{regdoc}},
\code{\link{regdocyr}}

Other Health: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docvig}},
\code{\link{regdoc}},
\code{\link{regdocyr}}
}
\concept{Doctors & Patients}
\concept{Health}
\keyword{variable}
