% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{docsat}
\alias{docsat}
\title{How satisfied r with office visit}
\source{
General Social Survey https://gss.norc.org
}
\description{
docsat
}
\details{
Question Based on what the doctor has just told you, how satisfied would you be with the office visit?  Would you rate the visit as...
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2755/vshow}{GSS Data Explorer page for \code{docsat}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 89 \tab 372 \tab 759 \tab 697 \tab 814 \tab 30 \tab 4 \tab 2765 \cr
   Total \tab 0 \tab 89 \tab 372 \tab 759 \tab 697 \tab 814 \tab 30 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{doccosts}},
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docswtch}},
\code{\link{insrchng}},
\code{\link{secopin}}
}
\concept{Doctors & Patients}
\keyword{variable}
