% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{docsat1}
\alias{docsat1}
\title{How satisfied r was with last doctor visit}
\source{
General Social Survey https://gss.norc.org
}
\description{
docsat1
}
\details{
Question How satisfied or dissatisfied were you with the treatment you received... When you last visited a doctor?
}
\section{Values}{

\itemize{
\item \verb{[1]} completely satisfied
\item \verb{[2]} very satisfied
\item \verb{[3]} fairly satisfied
\item \verb{[4]} neither satisfied nor dissatisfied
\item \verb{[5]} fairly dissatisfied
\item \verb{[6]} very dissatisfied
\item \verb{[7]} completely dissatisfied
\item \verb{[8]} does not apply
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7832/vshow}{GSS Data Explorer page for \code{docsat1}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab completely satisfied \tab very satisfied \tab fairly satisfied \tab neither satisfied nor dissatisfied \tab fairly dissatisfied \tab very dissatisfied \tab completely dissatisfied \tab 9 \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2382 \tab 224 \tab 382 \tab 314 \tab 78 \tab 66 \tab 21 \tab 29 \tab 30 \tab 11 \tab 3 \tab 4 \tab 3544 \cr
   Total \tab 2382 \tab 224 \tab 382 \tab 314 \tab 78 \tab 66 \tab 21 \tab 29 \tab 30 \tab 11 \tab 3 \tab 4 \tab 3544 \cr
}
}

\keyword{variable}
