% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{docvig}
\alias{docvig}
\title{Vignette about the doctor}
\source{
General Social Survey https://gss.norc.org
}
\description{
docvig
}
\details{
Question Vignette number given to respondent.
}
\section{Values}{

\itemize{
\item \verb{[1]} our physcian s group...(c4)
\item \verb{[2]} it use to be that physicans...(c5)
\item \verb{[3]} i can understand...(c6)
\item \verb{[4]} i understand that...(c7)
\item \verb{[5]} i have had a lot...(c8)
\item \verb{[6]} i usually don t order...(c9)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2737/vshow}{GSS Data Explorer page for \code{docvig}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab our physcian s group...(c4) \tab it use to be that physicans...(c5) \tab i can understand...(c6) \tab i understand that...(c7) \tab i have had a lot...(c8) \tab i usually don t order...(c9) \tab Total \cr
   2002 \tab - \tab 466 \tab 461 \tab 451 \tab 464 \tab 465 \tab 458 \tab 2765 \cr
   Total \tab 0 \tab 466 \tab 461 \tab 451 \tab 464 \tab 465 \tab 458 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{regdoc}},
\code{\link{regdocyr}}
}
\concept{Doctors & Patients}
\keyword{variable}
