% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{dofirst}
\alias{dofirst}
\title{Here are the things you said (#secl_name) should do...	 interviewer: read list}
\source{
General Social Survey https://gss.norc.org
}
\description{
dofirst
}
\details{
Question Here are the things you said (John/Juan/Mary/Maria) should do...
What is the first thing s/he should do?
}
\section{Values}{

\itemize{
\item \verb{[1]} talk to family and friends about it
\item \verb{[2]} talk to a minister, priest, rabbi or other religious leader
\item \verb{[3]} go to a general medical doctor for help
\item \verb{[4]} go to a psychiatrist for help
\item \verb{[5]} go to a therapist, or counselor, like a psychologist, social worker, or other mental health professional for help
\item \verb{[6]} go to a spiritual or a natural healer for help
\item \verb{[7]} join a self-help group where people with similar problems help each other
\item \verb{[8]} take non-prescription medication, like over the counter sleeping pills
\item \verb{[9]} take prescription medication
\item \verb{[10]} check into a mental hospital
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7094/vshow}{GSS Data Explorer page for \code{dofirst}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab talk to family and friends about it \tab talk to a minister, priest, rabbi or other religious leader \tab go to a general medical doctor for help \tab go to a psychiatrist for help \tab go to a therapist, or counselor, like a psychologist, social worker, or other mental health professional for help \tab go to a spiritual or a natural healer for help \tab join a self-help group where people with similar problems help each other \tab take prescription medication \tab check into a mental hospital \tab don't know \tab Total \cr
   2018 \tab 1214 \tab 679 \tab 26 \tab 209 \tab 42 \tab 104 \tab 2 \tab 37 \tab 2 \tab 29 \tab 4 \tab 2348 \cr
   Total \tab 1214 \tab 679 \tab 26 \tab 209 \tab 42 \tab 104 \tab 2 \tab 37 \tab 2 \tab 29 \tab 4 \tab 2348 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}
}
\concept{Mental Health}
\keyword{variable}
