% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{dofirst}
\alias{dofirst}
\title{Here are the things you said (#secl_name) should do...	 interviewer: read list}
\source{
General Social Survey https://gss.norc.org
}
\description{
dofirst
}
\details{
Question Here are the things you said (John/Juan/Mary/Maria) should do...
What is the first thing s/he should do?
}
\section{Values}{

\itemize{
\item \code{1} talk to family and friends about it
\item \code{2} talk to a minister, priest, rabbi or other religious leader
\item \code{3} go to a general medical doctor for help
\item \code{4} go to a psychiatrist for help
\item \code{5} go to a therapist, or counselor, like a psychologist, social worker, or other mental health professional for help
\item \code{6} go to a spiritual or a natural healer for help
\item \code{7} join a self-help group where people with similar problems help each other
\item \code{8} take non-prescription medication, like over the counter sleeping pills
\item \code{9} take prescription medication
\item \code{10} check into a mental hospital
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7094/vshow}{GSS Data Explorer page for \code{dofirst}}.
}

\seealso{
Other Mental Health: 
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}
}
\concept{Mental Health}
\keyword{variable}
