% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_420.R
\name{done911c}
\alias{done911c}
\title{Has r purchased things to make them safer}
\source{
General Social Survey https://gss.norc.org
}
\description{
done911c
}
\details{
Question (Do you know anyone who has done any of the following things because of terrorism since September 11th, 2001?)
Have you, or anyone you know purchased things to make them safer (gas masks, duct tape, things to make their house safer, etc.)?
FI PROBE: IF RESPONDENT ANSWERS ‘YES,’ ASK; ‘Was this you, someone you know, or both yourself and someone you know?
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes, respondent
\item \code{2} yes, someone respondent knows
\item \code{3} yes, both respondent and someone respondent knows
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3847/vshow}{GSS Data Explorer page for \code{done911c}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab yes, respondent \tab yes, someone respondent knows \tab yes, both respondent and someone respondent knows \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 1157 \tab 38 \tab 98 \tab 47 \tab 7 \tab 5 \tab 2023 \cr
   Total \tab 671 \tab 1157 \tab 38 \tab 98 \tab 47 \tab 7 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911a}},
\code{\link{done911b}},
\code{\link{eff911a}},
\code{\link{eff911b}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
