% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_422.R
\name{done911l}
\alias{done911l}
\title{Has r avoided tall buildings}
\source{
General Social Survey https://gss.norc.org
}
\description{
done911l
}
\details{
Question (Do you know anyone who has done any of the following things because of terrorism since September 11th, 2001?)
Have you, or anyone you know avoided tall buildings?
FI PROBE: IF RESPONDENT ANSWERS 'YES,' ASK; 'Was this you, someone you know, or both yourself and someone you know?
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes, respondent
\item \verb{[2]} yes, someone respondent knows
\item \verb{[3]} yes, both respondent and someone respondent knows
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3865/vshow}{GSS Data Explorer page for \code{done911l}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab yes, respondent \tab yes, someone respondent knows \tab yes, both respondent and someone respondent knows \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 1253 \tab 39 \tab 23 \tab 16 \tab 15 \tab 6 \tab 2023 \cr
   Total \tab 671 \tab 1253 \tab 39 \tab 23 \tab 16 \tab 15 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911m}},
\code{\link{eff911h}},
\code{\link{eff911i}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
