% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{doneelse}
\alias{doneelse}
\title{Specify what it is r has done}
\source{
General Social Survey https://gss.norc.org
}
\description{
doneelse
}
\details{
Question Please specify what it is you or someone you know has done.
}
\section{Values}{

\itemize{
\item \code{1} research and collect more information
\item \code{2} join military
\item \code{3} avoid possible targets of terror
\item \code{4} buy new insurance policy
\item \code{5} psycho therapy
\item \code{6} pray
\item \code{7} more vigilant to surrounding
\item \code{8} join secret service
\item \code{9} buy weapons
\item \code{10} vigilance and new assignments because they are already in military or secrete service
\item \code{11} terror-related traiing at work
\item \code{12} terror-related traiing in community
\item \code{13} build safe place
\item \code{14} deal with government
\item \code{15} talk about it with other people
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3871/vshow}{GSS Data Explorer page for \code{doneelse}}.
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911n}},
\code{\link{eff911m}},
\code{\link{eff911n}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
