% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{doubts2}
\alias{doubts2}
\title{Has personal suffering caused r doubt}
\source{
General Social Survey https://gss.norc.org
}
\description{
doubts2
}
\details{
Question How often have these problems caused doubts about your religious faith- often, sometimes, or never.
B. Personal suffering
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1096/vshow}{GSS Data Explorer page for \code{doubts2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 130 \tab 650 \tab 657 \tab 26 \tab 18 \tab 1481 \cr
   1998 \tab 1387 \tab 126 \tab 527 \tab 743 \tab 37 \tab 12 \tab 2832 \cr
   Total \tab 1387 \tab 256 \tab 1177 \tab 1400 \tab 63 \tab 30 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
