% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_459.R
\name{down1a}
\alias{down1a}
\title{Who does r turn to first for help when depressed}
\source{
General Social Survey https://gss.norc.org
}
\description{
down1a
}
\details{
Question Now suppose you felt just a bit down or depressed, and you wanted to talk about it.
A. Who would you turn to first for help?
}
\section{Values}{

\itemize{
\item \code{1} husband, wife, partner
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} son
\item \code{6} sister
\item \code{7} brother
\item \code{8} other blood relative
\item \code{9} other in-law relative
\item \code{10} close friend
\item \code{11} neighbor
\item \code{12} someone you work with
\item \code{13} priest or member of the clergy
\item \code{14} family doctor
\item \code{15} a psychologist or other professional counselor
\item \code{16} a self-help group
\item \code{17} someone else
\item \code{18} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4218/vshow}{GSS Data Explorer page for \code{down1a}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1a}},
\code{\link{borrow2a}},
\code{\link{down2a}},
\code{\link{grpchrty}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{helphwrk}},
\code{\link{sick1a}},
\code{\link{sick2a}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
