% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_299.R
\name{downfeel}
\alias{downfeel}
\title{Treatment for downheartedness improve feelings}
\source{
General Social Survey https://gss.norc.org
}
\description{
downfeel
}
\details{
Question As you know, when people go to doctors, they may hold different expectations of what should result from their treatment. I’m going to describe a health problem. Please tell me how much you, yourself, would expect to get each of the following:
E. Would you definitely expect, probably expect, probably not expect, or definitely not expect to feel better about yourself as a person?
}
\section{Values}{

\itemize{
\item \code{1} definitely expect
\item \code{2} probably expect
\item \code{3} probably not expect
\item \code{4} definitely not expect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2714/vshow}{GSS Data Explorer page for \code{downfeel}}.
}

\seealso{
Other Health Status: 
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotqol}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}}

Other Health: 
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotqol}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painfeel}},
\code{\link{painqol}},
\code{\link{painrely}}
}
\concept{Health}
\concept{Health Status}
\keyword{variable}
