% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{draft}
\alias{draft}
\title{Return to the draft?}
\source{
General Social Survey https://gss.norc.org
}
\description{
draft
}
\details{
Question Do you think we should return to a military draft at this time, or should we continue to rely on volunteers?
}
\section{Values}{

\itemize{
\item \code{1} draft
\item \code{2} volunteers
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/802/vshow}{GSS Data Explorer page for \code{draft}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab draft \tab volunteers \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 719 \tab 1012 \tab 122 \tab 7 \tab 1860 \cr
   1983 \tab - \tab 458 \tab 1070 \tab 66 \tab 5 \tab 1599 \cr
   1984 \tab - \tab 329 \tab 1058 \tab 64 \tab 22 \tab 1473 \cr
   2006 \tab 2992 \tab 237 \tab 1244 \tab 35 \tab 2 \tab 4510 \cr
   Total \tab 2992 \tab 1743 \tab 4384 \tab 287 \tab 36 \tab 9442 \cr
}
}

\seealso{
Other Core: 
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}},
\code{\link{fehlpmil}}
}
\concept{Core}
\keyword{variable}
