% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{draftfe}
\alias{draftfe}
\title{If return to draft, draft women?}
\source{
General Social Survey https://gss.norc.org
}
\description{
draftfe
}
\details{
Question Do you think we should return to a military draft at this time, or should we continue to rely on volunteers?
A. IF DRAFT: If we should return to a military draft at this time, should young women be drafted as well as young men, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/803/vshow}{GSS Data Explorer page for \code{draftfe}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1134 \tab 383 \tab 314 \tab 19 \tab 10 \tab 1860 \cr
   1983 \tab 1136 \tab 212 \tab 234 \tab 11 \tab 6 \tab 1599 \cr
   1984 \tab 1122 \tab 160 \tab 161 \tab 7 \tab 23 \tab 1473 \cr
   Total \tab 3392 \tab 755 \tab 709 \tab 37 \tab 39 \tab 4932 \cr
}
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}},
\code{\link{fehlpmil}}
}
\concept{Core}
\keyword{variable}
