% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{draftfem}
\alias{draftfem}
\title{If emergency, draft women also?}
\source{
General Social Survey https://gss.norc.org
}
\description{
draftfem
}
\details{
Question Do you think we should return to a military draft at this time, or should we continue to rely on volunteers?
C. IF DRAFT ON "B": If we should return to a military draft in a national emergency, should young women be drafted as well as young men, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/805/vshow}{GSS Data Explorer page for \code{draftfem}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not \tab don't know \tab no answer \tab Total \cr
   1982 \tab 977 \tab 430 \tab 398 \tab 30 \tab 25 \tab 1860 \cr
   1983 \tab 645 \tab 460 \tab 436 \tab 29 \tab 29 \tab 1599 \cr
   1984 \tab 539 \tab 451 \tab 408 \tab 22 \tab 53 \tab 1473 \cr
   Total \tab 2161 \tab 1341 \tab 1242 \tab 81 \tab 107 \tab 4932 \cr
}
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}},
\code{\link{fehlpmil}}
}
\concept{Core}
\keyword{variable}
