% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{drinkday}
\alias{drinkday}
\title{How many drinks r have on a day when drinking}
\source{
General Social Survey https://gss.norc.org
}
\description{
drinkday
}
\details{
Question Again, as you think back over the last 12 months, about how many drinks would you have on a typical day when you  drank?
PROBE IF NECESSARY:  Would it have been 5 or more?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3099/vshow}{GSS Data Explorer page for \code{drinkday}}.
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkmax}},
\code{\link{drinkyr}},
\code{\link{evdrink}}
}
\concept{Women & Alcohol}
\keyword{variable}
