% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{drinkmax}
\alias{drinkmax}
\title{Largest number of drinks in a day last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
drinkmax
}
\details{
Question Think of all kinds of alcoholic beverages combined, that is, any combination of cans, bottles or glasses of beer; glasses of wine; or drinks containing liquor of any kind. During the last 12 months, what is the largest number of drinks you had on any single day?  Was it:
}
\section{Values}{

\itemize{
\item \code{0} did not drink at all in the last 12 months
\item \code{1} at least a sip, but less than one full drink
\item \code{2} at least 1, but less than 3 drinks
\item \code{3} at least 3, but less than 5 drinks
\item \code{4} at least 5, but less than 8 drinks
\item \code{5} at least 8, but less than 12 drinks
\item \code{6} at least 12, but less than 20 drinks
\item \code{7} 20 or more drinks in a single day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3101/vshow}{GSS Data Explorer page for \code{drinkmax}}.
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkday}},
\code{\link{drinkyr}},
\code{\link{evdrink}}

Other Drinking: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkday}},
\code{\link{drinkyr}},
\code{\link{evdrink}}
}
\concept{Drinking}
\concept{Women & Alcohol}
\keyword{variable}
