% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{drinkmin}
\alias{drinkmin}
\title{How often at least a sip but less than one full drink}
\source{
General Social Survey https://gss.norc.org
}
\description{
drinkmin
}
\details{
Question F. During the last 12 months, how often did you have at least a sip, but less than one full drink of any kind of alcoholic beverage in a single day, that is, beer, wine, or a drink containing liquor of any kind. Was it:
}
\section{Values}{

\itemize{
\item \code{1} never in the last 12 months
\item \code{2} once in the last 12 months
\item \code{3} twice in the last 12 months
\item \code{4} three to six times in the last 12 months
\item \code{5} seven to eleven times in the last 12 months
\item \code{6} one to three times a month
\item \code{7} once or twice a week
\item \code{8} three or four times a week
\item \code{9} every day or nearly every day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3107/vshow}{GSS Data Explorer page for \code{drinkmin}}.
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink1}},
\code{\link{drink3}},
\code{\link{drink5}}

Other Drinking: 
\code{\link{drink1}},
\code{\link{drink3}},
\code{\link{drink5}}
}
\concept{Drinking}
\concept{Women & Alcohol}
\keyword{variable}
