% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_345.R
\name{drinkyr}
\alias{drinkyr}
\title{How regularly r drink over last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
drinkyr
}
\details{
Question Thinking back over the last 12 months, about how regularly did you drink alcoholic beverages?  Would you say that it was . . . READ OPTIONS DRINKYR
}
\section{Values}{

\itemize{
\item \code{1} never in those 12 months
\item \code{2} 1 to 3 times in 12 months
\item \code{3} 4 to 7 times in 12 months
\item \code{4} 8 to 11 times in 12 months
\item \code{5} 1 to 3 times a month
\item \code{6} once or twice a week
\item \code{7} 3 to 4 times a week
\item \code{8} 5 times a week or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3098/vshow}{GSS Data Explorer page for \code{drinkyr}}.
}

\seealso{
Other Women & Alcohol: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkday}},
\code{\link{drinkmax}},
\code{\link{evdrink}}

Other Drinking: 
\code{\link{drink12}},
\code{\link{drink6up}},
\code{\link{drink8}},
\code{\link{drinkday}},
\code{\link{drinkmax}},
\code{\link{evdrink}}
}
\concept{Drinking}
\concept{Women & Alcohol}
\keyword{variable}
