% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{dwelcity}
\alias{dwelcity}
\title{Dwelling compared to others in city}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwelcity
}
\details{
Question B. Compared to houses/apartments in the city/town/county, would you say the house was . . .
}
\section{Values}{

\itemize{
\item \code{1} far above average
\item \code{2} above average
\item \code{3} average
\item \code{4} below average
\item \code{5} far below average
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5293/vshow}{GSS Data Explorer page for \code{dwelcity}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab far above average \tab above average \tab average \tab below average \tab far below average \tab no answer \tab Total \cr
   1982 \tab - \tab 37 \tab 318 \tab 982 \tab 382 \tab 72 \tab 69 \tab 1860 \cr
   1983 \tab - \tab 43 \tab 312 \tab 872 \tab 300 \tab 44 \tab 28 \tab 1599 \cr
   1984 \tab - \tab 34 \tab 272 \tab 810 \tab 273 \tab 38 \tab 46 \tab 1473 \cr
   Total \tab 0 \tab 114 \tab 902 \tab 2664 \tab 955 \tab 154 \tab 143 \tab 4932 \cr
}
}

\seealso{
Other Residence: 
\code{\link{dwelling}},
\code{\link{dwelngh}}
}
\concept{Residence}
\keyword{variable}
