% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{dwell5}
\alias{dwell5}
\title{In last 5 yrs, r has been prevented from moving bc landlord believed r was gay/l}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwell5
}
\details{
Question In the last five years, have you been forced to move or prevented from buying or moving into a house or apartment  because a landlord or realtor believed you were gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3927/vshow}{GSS Data Explorer page for \code{dwell5}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}

Other Dwelling (R's): 
\code{\link{evdwell}}

Other Gays And Lesbians: 
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}

Other Residence: 
\code{\link{evdwell}}
}
\concept{Dwelling (R's)}
\concept{Gays And Lesbians}
\concept{Residence}
\concept{Sexual Orientation}
\keyword{variable}
