% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{dwell5}
\alias{dwell5}
\title{In last 5 yrs, r has been prevented from moving bc landlord believed r was gay/l}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwell5
}
\details{
Question In the last five years, have you been forced to move or prevented from buying or moving into a house or apartment  because a landlord or realtor believed you were gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3927/vshow}{GSS Data Explorer page for \code{dwell5}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2008 \tab 1995 \tab 1 \tab 1 \tab 26 \tab 2023 \cr
   2012 \tab 1925 \tab 1 \tab 2 \tab 46 \tab 1974 \cr
   Total \tab 3920 \tab 2 \tab 3 \tab 72 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
