% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{dwellpre}
\alias{dwellpre}
\title{Type of structure in which the respondent lives}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwellpre
}
\details{
Question SELECT ONE FROM THE LIST BELOW WHICH BEST DESCRIBES THE TYPE OF STRUCTURE IN WHICH THE RESPONDENT LIVES
}
\section{Values}{

\itemize{
\item \code{1} trailer
\item \code{2} detached single family house
\item \code{3} 2-family house, 2 units side-by-side
\item \code{4} 2-family house, 2 units one above the other
\item \code{5} detached 3-4 family house
\item \code{6} row house (3 or more units in an attached row)
\item \code{7} apartment house (5 or more units, 3 stories or less)
\item \code{8} apartment house (5 or more units, 4 stories or more)
\item \code{9} apartment in a partly commercial structure
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6654/vshow}{GSS Data Explorer page for \code{dwellpre}}.
}

\keyword{variable}
