% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{dwelngh}
\alias{dwelngh}
\title{Dwelling compared to others in nbrhd}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwelngh
}
\details{
Question A. Compared to houses/apartments in the neighborhood, would you say the house/ apartment was . . .
}
\section{Values}{

\itemize{
\item \code{1} far above average
\item \code{2} above average
\item \code{3} average
\item \code{4} below average
\item \code{5} far below average
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5292/vshow}{GSS Data Explorer page for \code{dwelngh}}.
}

\seealso{
Other Residence: 
\code{\link{dwelcity}},
\code{\link{dwelling}}

Other Dwelling (R's): 
\code{\link{dwelcity}},
\code{\link{dwelling}}
}
\concept{Dwelling (R's)}
\concept{Residence}
\keyword{variable}
