% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_578.R
\name{dwelown16}
\alias{dwelown16}
\title{Did rs family own or rent home when r was age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
dwelown16
}
\details{
Question When you were 16 years old, did your family own your own home, pay rent, or something else?
}
\section{Values}{

\itemize{
\item \code{1} owned or was buying
\item \code{2} paid rent
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7299/vshow}{GSS Data Explorer page for \code{dwelown16}}.
}

\keyword{variable}
