% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{earndes}
\alias{earndes}
\title{R deserves earnings}
\source{
General Social Survey https://gss.norc.org
}
\description{
earndes
}
\details{
Question Would you say that you earn:
}
\section{Values}{

\itemize{
\item \code{1} much less than i deserve
\item \code{2} less than i deserve
\item \code{3} what i deserve
\item \code{4} more than i deserve
\item \code{5} much more than i deserve
\item \code{6} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4323/vshow}{GSS Data Explorer page for \code{earndes}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab much less than i deserve \tab less than i deserve \tab what i deserve \tab more than i deserve \tab much more than i deserve \tab never worked \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 188 \tab 471 \tab 407 \tab 34 \tab 4 \tab 17 \tab 68 \tab 83 \tab 2817 \cr
   Total \tab 1545 \tab 188 \tab 471 \tab 407 \tab 34 \tab 4 \tab 17 \tab 68 \tab 83 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
