% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{earnmore}
\alias{earnmore}
\title{Who earns more money?}
\source{
General Social Survey https://gss.norc.org
}
\description{
earnmore
}
\details{
Question Who earns more money?
}
\section{Values}{

\itemize{
\item \code{1} i earn much more than my spouse
\item \code{2} i earn somewhat more than my spouse
\item \code{3} we earn about the same amount
\item \code{4} my spouser-partner earns somewhat more
\item \code{5} my spouse-partner earns much more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1663/vshow}{GSS Data Explorer page for \code{earnmore}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab i earn much more than my spouse \tab i earn somewhat more than my spouse \tab we earn about the same amount \tab my spouser-partner earns somewhat more \tab my spouse-partner earns much more \tab no answer \tab Total \cr
   1996 \tab 2124 \tab 278 \tab 66 \tab 81 \tab 130 \tab 200 \tab 25 \tab 2904 \cr
   Total \tab 2124 \tab 278 \tab 66 \tab 81 \tab 130 \tab 200 \tab 25 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}
}
\concept{Gender}
\keyword{variable}
