% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{earnshh}
\alias{earnshh}
\title{Hubby or wife earns more dollars}
\source{
General Social Survey https://gss.norc.org
}
\description{
earnshh
}
\details{
Question B. Considering all sources of income, between you and your spouse/partner, who has the higher income?
}
\section{Values}{

\itemize{
\item \code{1} spouse has no income
\item \code{2} i have much higher income
\item \code{3} i have higher income
\item \code{4} we have about the same income
\item \code{5} spouse has a higher income
\item \code{6} spouse partner has much higher income
\item \code{7} i have no income
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4456/vshow}{GSS Data Explorer page for \code{earnshh}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab spouse has no income \tab i have much higher income \tab i have higher income \tab we have about the same income \tab spouse has a higher income \tab spouse partner has much higher income \tab i have no income \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2058 \tab 65 \tab 127 \tab 100 \tab 75 \tab 198 \tab 70 \tab 29 \tab 18 \tab 25 \tab 2765 \cr
   2012 \tab 1274 \tab 58 \tab 112 \tab 142 \tab 55 \tab 236 \tab 54 \tab 34 \tab 2 \tab 7 \tab 1974 \cr
   Total \tab 3332 \tab 123 \tab 239 \tab 242 \tab 130 \tab 434 \tab 124 \tab 63 \tab 20 \tab 32 \tab 4739 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
