% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{earnsmor}
\alias{earnsmor}
\title{Hubby or wife earns more $}
\source{
General Social Survey https://gss.norc.org
}
\description{
earnsmor
}
\details{
Question Who earns more money?
}
\section{Values}{

\itemize{
\item \code{1} husband, much more
\item \code{2} husband, bit more
\item \code{3} same for both
\item \code{4} wife, bit more
\item \code{5} wife, much more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4400/vshow}{GSS Data Explorer page for \code{earnsmor}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabit}},
\code{\link{divhubby}},
\code{\link{evdiv}},
\code{\link{mawork14}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\concept{Income}
\keyword{variable}
