% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{earthsun}
\alias{earthsun}
\title{Sci knowledge: the earth goes around the sun}
\source{
General Social Survey https://gss.norc.org
}
\description{
earthsun
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don’t know or aren’t sure, just tell me so,  and we will skip to the next question. Remember true, false, or don’t know.
J. Now, does the Earth go around the Sun, or does the Sun go around the Earth?
}
\section{Values}{

\itemize{
\item \code{1} earth around sun
\item \code{2} sun around earth
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3343/vshow}{GSS Data Explorer page for \code{earthsun}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}

Other Knowledge: 
\code{\link{condrift}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{solarrev}}

Other Science: 
\code{\link{condrift}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{solarrev}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
