% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{easyget}
\alias{easyget}
\title{How difficult was case to get}
\source{
General Social Survey https://gss.norc.org
}
\description{
easyget
}
\details{
Question How difficult was case to get?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} somewhat easy
\item \code{3} somewhat difficult
\item \code{4} very difficult
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5564/vshow}{GSS Data Explorer page for \code{easyget}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very easy \tab somewhat easy \tab somewhat difficult \tab very difficult \tab no answer \tab Total \cr
   1998 \tab - \tab 687 \tab 796 \tab 815 \tab 494 \tab 40 \tab 2832 \cr
   2000 \tab - \tab 733 \tab 827 \tab 783 \tab 413 \tab 61 \tab 2817 \cr
   Total \tab 0 \tab 1420 \tab 1623 \tab 1598 \tab 907 \tab 101 \tab 5649 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}},
\code{\link{mode}()}
}
\concept{Paradata}
\keyword{variable}
