% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{eatmeat}
\alias{eatmeat}
\title{In a typical week, on how many days do you eat beef, lamb, or products that cont}
\source{
General Social Survey https://gss.norc.org
}
\description{
eatmeat
}
\details{
Question In a typical week, on how many days do you eat beef, lamb, or products that contain them?
}
\section{Values}{

\itemize{
\item \verb{[0]} none
\item \verb{[7]} 7 days
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7499/vshow}{GSS Data Explorer page for \code{eatmeat}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab none \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 days \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 214 \tab 315 \tab 365 \tab 342 \tab 206 \tab 163 \tab 67 \tab 123 \tab 34 \tab 192 \tab 8 \tab 4032 \cr
   Total \tab 2003 \tab 214 \tab 315 \tab 365 \tab 342 \tab 206 \tab 163 \tab 67 \tab 123 \tab 34 \tab 192 \tab 8 \tab 4032 \cr
}
}

\keyword{variable}
