% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{eatmeat}
\alias{eatmeat}
\title{In a typical week, on how many days do you eat beef, lamb, or products that cont}
\source{
General Social Survey https://gss.norc.org
}
\description{
eatmeat
}
\details{
Question In a typical week, on how many days do you eat beef, lamb, or products that contain them?
}
\section{Values}{

\itemize{
\item \code{0} none
\item \code{7} 7 days
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7499/vshow}{GSS Data Explorer page for \code{eatmeat}}.
}

\keyword{variable}
