% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{econgrn}
\alias{econgrn}
\title{Economic progress dependent on environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
econgrn
}
\details{
Question How much do you agree or disagree with each of these statements?
C. Economic progress in America will slow down unless we look after the environment better.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4789/vshow}{GSS Data Explorer page for \code{econgrn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 126 \tab 438 \tab 353 \tab 171 \tab 23 \tab 67 \tab 98 \tab 2817 \cr
   2010 \tab 614 \tab 122 \tab 621 \tab 293 \tab 268 \tab 34 \tab 85 \tab 7 \tab 2044 \cr
   Total \tab 2155 \tab 248 \tab 1059 \tab 646 \tab 439 \tab 57 \tab 152 \tab 105 \tab 4861 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{excldimm}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{ldcgrn}},
\code{\link{nukeacc}}
}
\concept{ISSP Environment}
\keyword{variable}
