% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_175.R
\name{econpast}
\alias{econpast}
\title{Has national economy improved last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
econpast
}
\details{
Question How about the economy. Would you say that over the past year the nation's economy has gotten better, stayed about the same, or gotten worse?
}
\section{Values}{

\itemize{
\item \code{1} gotten better
\item \code{2} stayed the same
\item \code{3} gotten worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1641/vshow}{GSS Data Explorer page for \code{econpast}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{hspgovt}},
\code{\link{obrespct}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Economy}
\concept{Multiculturalism}
\keyword{variable}
