% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{econsup}
\alias{econsup}
\title{Main source of economic support}
\source{
General Social Survey https://gss.norc.org
}
\description{
econsup
}
\details{
Question What is your main source of economic support?
}
\section{Values}{

\itemize{
\item \code{1} pnsion pri/st/ss
\item \code{2} unemploy benfits
\item \code{3} spouse/partner
\item \code{4} othr famly mmbrs
\item \code{5} soc assis/welfre
\item \code{6} occasional work
\item \code{7} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4605/vshow}{GSS Data Explorer page for \code{econsup}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab pnsion pri/st/ss \tab unemploy benfits \tab spouse/partner \tab othr famly mmbrs \tab soc assis/welfre \tab occasional work \tab other \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2460 \tab 152 \tab 8 \tab 66 \tab 28 \tab 29 \tab 4 \tab 43 \tab 2 \tab 40 \tab 2832 \cr
   2006 \tab 4008 \tab 173 \tab 19 \tab 119 \tab 30 \tab 68 \tab 8 \tab 78 \tab - \tab 7 \tab 4510 \cr
   Total \tab 6468 \tab 325 \tab 27 \tab 185 \tab 58 \tab 97 \tab 12 \tab 121 \tab 2 \tab 47 \tab 7342 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{rimpskls}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}
}
\concept{Work Orientation}
\keyword{variable}
