% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{econsys}
\alias{econsys}
\title{Best-needs replaced}
\source{
General Social Survey https://gss.norc.org
}
\description{
econsys
}
\details{
Question On the whole, do you think our economic system is...
}
\section{Values}{

\itemize{
\item \code{1} the best possible
\item \code{2} ok but need some tinkering
\item \code{3} need fundamental changes
\item \code{4} need to be replaced
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1879/vshow}{GSS Data Explorer page for \code{econsys}}.
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}

Other Economy: 
\code{\link{reqinfo}}
}
\concept{Economy}
\concept{Markets}
\keyword{variable}
