% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{eddone}
\alias{eddone}
\title{Young should complete formal schooling}
\source{
General Social Survey https://gss.norc.org
}
\description{
eddone
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important is it for them to be...
Completed their formal schooling
}
\section{Values}{

\itemize{
\item \code{1} extremely important
\item \code{2} quite important
\item \code{3} somewhat important
\item \code{4} not too important
\item \code{5} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2900/vshow}{GSS Data Explorer page for \code{eddone}}.
}

\seealso{
Other Adult Transitions: 
\code{\link{eddone1}},
\code{\link{finind}},
\code{\link{finind1}},
\code{\link{ftwork}},
\code{\link{ftwork1}},
\code{\link{ownhh}},
\code{\link{ownhh1}}

Other Education: 
\code{\link{eddone1}}
}
\concept{Adult Transitions}
\concept{Education}
\keyword{variable}
