% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{educ1}
\alias{educ1}
\title{Education level of person #1}
\source{
General Social Survey https://gss.norc.org
}
\description{
educ1
}
\details{
Question This card (CARD BH) lists general levels of education. As far as you know, what is (NAME)''s highest level of education?
PROBE:  Your best guess. RECORD VERBATIM IF NOT CODEABLE. ASK FOR EACH NAME.
\enumerate{
\item Education of First Person
}
}
\section{Values}{

\itemize{
\item \verb{[0]} 1-6 years
\item \verb{[1]} 7-9 years
\item \verb{[2]} 10-12 years
\item \verb{[3]} h.s. grad
\item \verb{[4]} some college
\item \verb{[5]} asso. degree
\item \verb{[6]} bach. degree
\item \verb{[7]} grad or prof
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/940/vshow}{GSS Data Explorer page for \code{educ1}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1-6 years \tab 7-9 years \tab 10-12 years \tab h.s. grad \tab some college \tab asso. degree \tab bach. degree \tab grad or prof \tab don't know \tab no answer \tab Total \cr
   1985 \tab 136 \tab 17 \tab 84 \tab 168 \tab 463 \tab 231 \tab 63 \tab 193 \tab 142 \tab 25 \tab 12 \tab 1534 \cr
   2004 \tab 1696 \tab 8 \tab 32 \tab 107 \tab 297 \tab 169 \tab 71 \tab 208 \tab 155 \tab 17 \tab 52 \tab 2812 \cr
   Total \tab 1832 \tab 25 \tab 116 \tab 275 \tab 760 \tab 400 \tab 134 \tab 401 \tab 297 \tab 42 \tab 64 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}}
}
\concept{Social Networks}
\keyword{variable}
