% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{educ3}
\alias{educ3}
\title{Education level of person #3}
\source{
General Social Survey https://gss.norc.org
}
\description{
educ3
}
\details{
Question This card (CARD BH) lists general levels of education. As far as you know, what is (NAME)''s highest level of education?
PROBE:  Your best guess. RECORD VERBATIM IF NOT CODEABLE. ASK FOR EACH NAME.
3. Education of Third Person
}
\section{Values}{

\itemize{
\item \code{0} 1-6 years
\item \code{1} 7-9 years
\item \code{2} 10-12 years
\item \code{3} h.s. grad
\item \code{4} some college
\item \code{5} asso. degree
\item \code{6} bach. degree
\item \code{7} grad or prof
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/942/vshow}{GSS Data Explorer page for \code{educ3}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1-6 years \tab 7-9 years \tab 10-12 years \tab h.s. grad \tab some college \tab asso. degree \tab bach. degree \tab grad or prof \tab don't know \tab no answer \tab Total \cr
   1985 \tab 599 \tab 13 \tab 46 \tab 82 \tab 322 \tab 157 \tab 34 \tab 128 \tab 113 \tab 27 \tab 13 \tab 1534 \cr
   2004 \tab 2240 \tab 3 \tab 8 \tab 36 \tab 122 \tab 86 \tab 41 \tab 125 \tab 94 \tab 10 \tab 47 \tab 2812 \cr
   Total \tab 2839 \tab 16 \tab 54 \tab 118 \tab 444 \tab 243 \tab 75 \tab 253 \tab 207 \tab 37 \tab 60 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ4}},
\code{\link{educ5}}
}
\concept{Social Networks}
\keyword{variable}
