% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{educ4}
\alias{educ4}
\title{Education level of person #4}
\source{
General Social Survey https://gss.norc.org
}
\description{
educ4
}
\details{
Question This card (CARD BH) lists general levels of education. As far as you know, what is (NAME)''s highest level of education?
PROBE:  Your best guess. RECORD VERBATIM IF NOT CODEABLE. ASK FOR EACH NAME.
4. Education of Fourth Person
}
\section{Values}{

\itemize{
\item \code{0} 1-6 years
\item \code{1} 7-9 years
\item \code{2} 10-12 years
\item \code{3} h.s. grad
\item \code{4} some college
\item \code{5} asso. degree
\item \code{6} bach. degree
\item \code{7} grad or prof
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/943/vshow}{GSS Data Explorer page for \code{educ4}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1-6 years \tab 7-9 years \tab 10-12 years \tab h.s. grad \tab some college \tab asso. degree \tab bach. degree \tab grad or prof \tab don't know \tab no answer \tab Total \cr
   1985 \tab 920 \tab 15 \tab 24 \tab 52 \tab 184 \tab 106 \tab 23 \tab 117 \tab 67 \tab 16 \tab 10 \tab 1534 \cr
   2004 \tab 2472 \tab - \tab 1 \tab 16 \tab 73 \tab 47 \tab 17 \tab 84 \tab 49 \tab 7 \tab 46 \tab 2812 \cr
   Total \tab 3392 \tab 15 \tab 25 \tab 68 \tab 257 \tab 153 \tab 40 \tab 201 \tab 116 \tab 23 \tab 56 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ5}}
}
\concept{Social Networks}
\keyword{variable}
