% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_103.R
\name{educ4}
\alias{educ4}
\title{Education level of person #4}
\source{
General Social Survey https://gss.norc.org
}
\description{
educ4
}
\details{
Question This card (CARD BH) lists general levels of education. As far as you know, what is (NAME)''s highest level of education?
PROBE:  Your best guess. RECORD VERBATIM IF NOT CODEABLE. ASK FOR EACH NAME.
4. Education of Fourth Person
}
\section{Values}{

\itemize{
\item \code{0} 1-6 years
\item \code{1} 7-9 years
\item \code{2} 10-12 years
\item \code{3} h.s. grad
\item \code{4} some college
\item \code{5} asso. degree
\item \code{6} bach. degree
\item \code{7} grad or prof
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/943/vshow}{GSS Data Explorer page for \code{educ4}}.
}

\seealso{
Other Social Networks: 
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ5}}

Other Education: 
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ5}}

Other Social Networks: 
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ5}}
}
\concept{Education}
\concept{Social Networks}
\keyword{variable}
