% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{eductech}
\alias{eductech}
\title{Are educated or uneducated benefiting more from digital tech}
\source{
General Social Survey https://gss.norc.org
}
\description{
eductech
}
\details{
Question Do you think that in America the highly educated or the less well educated are benefitting more, or both equally, or neither, from the use of the Internet and digital technology?
}
\section{Values}{

\itemize{
\item \verb{[1]} the highly educated are benefitting more than the less well educated
\item \verb{[2]} the highly and the less well educated are benefitting equally
\item \verb{[3]} the less well educated are benefitting more than the highly educated
\item \verb{[4]} neither highly educated nor less well educated are benefitting
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8106/vshow}{GSS Data Explorer page for \code{eductech}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab the highly educated are benefitting more than the less well educated \tab the highly and the less well educated are benefitting equally \tab the less well educated are benefitting more than the highly educated \tab neither highly educated nor less well educated are benefitting \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 714 \tab 407 \tab 108 \tab 140 \tab 241 \tab 1676 \tab 6 \tab 17 \tab 3309 \cr
   Total \tab 0 \tab 714 \tab 407 \tab 108 \tab 140 \tab 241 \tab 1676 \tab 6 \tab 17 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{aidrive}},
\code{\link{aimed}},
\code{\link{aiworry}},
\code{\link{classtech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
