% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{eductech}
\alias{eductech}
\title{Are educated or uneducated benefiting more from digital tech}
\source{
General Social Survey https://gss.norc.org
}
\description{
eductech
}
\details{
Question Do you think that in America the highly educated or the less well educated are benefitting more, or both equally, or neither, from the use of the Internet and digital technology?
}
\section{Values}{

\itemize{
\item \code{1} the highly educated are benefitting more than the less well educated
\item \code{2} the highly and the less well educated are benefitting equally
\item \code{3} the less well educated are benefitting more than the highly educated
\item \code{4} neither highly educated nor less well educated are benefitting
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8106/vshow}{GSS Data Explorer page for \code{eductech}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{classtech}},
\code{\link{dataprot}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{trcong}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
