% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_420.R
\name{eff911b}
\alias{eff911b}
\title{How effective r thinks stockpiling supplies}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911b
}
\details{
Question How effective do you think stockpiling supplies (food, water, antibiotics, etc.) is for people dealing with terrorism?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 not at all effective
\item \verb{[2]} 2
\item \verb{[3]} 3
\item \verb{[4]} 4
\item \verb{[5]} 5 extremely effective
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3846/vshow}{GSS Data Explorer page for \code{eff911b}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 not at all effective \tab 2 \tab 3 \tab 4 \tab 5 extremely effective \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 272 \tab 166 \tab 401 \tab 164 \tab 275 \tab 69 \tab 5 \tab 2023 \cr
   Total \tab 671 \tab 272 \tab 166 \tab 401 \tab 164 \tab 275 \tab 69 \tab 5 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911a}},
\code{\link{done911b}},
\code{\link{done911c}},
\code{\link{eff911a}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
