% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_421.R
\name{eff911c}
\alias{eff911c}
\title{How effective r thinks purchasing things to make them safer}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911c
}
\details{
Question How effective do you think purchasing things to make them safer (gas masks, duct tape, things to make their house safer, etc.) is for people dealing with terrorism? (Would you say 1, not at all effective, 5, extremely effective, or you may use any number in between?)
}
\section{Values}{

\itemize{
\item \code{1} 1 not at all effective
\item \code{2} 2
\item \code{3} 3
\item \code{4} 4
\item \code{5} 5 extremely effective
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3848/vshow}{GSS Data Explorer page for \code{eff911c}}.
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911d}},
\code{\link{done911e}},
\code{\link{done911f}},
\code{\link{done911g}},
\code{\link{done911h}},
\code{\link{eff911d}},
\code{\link{eff911e}},
\code{\link{eff911f}},
\code{\link{eff911g}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
