% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_422.R
\name{eff911i}
\alias{eff911i}
\title{How effective r thinks changing mail handling procedures}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911i
}
\details{
Question How effective do you think changing mail handling procedures is for people dealing with terrorism? (Would you say 1, not at all effective, 5, extremely effective, or you may use any number in between?)
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 not at all effective
\item \verb{[2]} 2
\item \verb{[3]} 3
\item \verb{[4]} 4
\item \verb{[5]} 5 extremely effective
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3860/vshow}{GSS Data Explorer page for \code{eff911i}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 not at all effective \tab 2 \tab 3 \tab 4 \tab 5 extremely effective \tab don't know \tab no answer \tab Total \cr
   2008 \tab 671 \tab 634 \tab 140 \tab 263 \tab 88 \tab 131 \tab 90 \tab 6 \tab 2023 \cr
   Total \tab 671 \tab 634 \tab 140 \tab 263 \tab 88 \tab 131 \tab 90 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911i}},
\code{\link{done911j}},
\code{\link{done911k}},
\code{\link{done911l}},
\code{\link{done911m}},
\code{\link{eff911h}},
\code{\link{eff911j}},
\code{\link{eff911k}},
\code{\link{eff911l}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
