% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{eff911m}
\alias{eff911m}
\title{How effective r thinks avoiding national landmarks}
\source{
General Social Survey https://gss.norc.org
}
\description{
eff911m
}
\details{
Question How effective do you think avoiding national landmarks is for people dealing with terrorism? (Would you say 1, not at all effective, 5, extremely effective, or you may use any number in between?)
}
\section{Values}{

\itemize{
\item \code{1} 1 not at all effective
\item \code{2} 2
\item \code{3} 3
\item \code{4} 4
\item \code{5} 5 extremely effective
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3868/vshow}{GSS Data Explorer page for \code{eff911m}}.
}

\seealso{
Other Terrorism Preparedness: 
\code{\link{done911n}},
\code{\link{doneelse}},
\code{\link{eff911n}}
}
\concept{Terrorism Preparedness}
\keyword{variable}
