% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{effctsup}
\alias{effctsup}
\title{Supervisor effective solve work/personal conflicts}
\source{
General Social Survey https://gss.norc.org
}
\description{
effctsup
}
\details{
Question How effectively does your supervisor work with employees to creatively solve conflicts between work and personal or family life…
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all effective
\item \verb{[2]} a little
\item \verb{[3]} somewhat
\item \verb{[4]} very
\item \verb{[5]} extremely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5506/vshow}{GSS Data Explorer page for \code{effctsup}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all effective \tab a little \tab somewhat \tab very \tab extremely \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1505 \tab 93 \tab 158 \tab 288 \tab 339 \tab 125 \tab 24 \tab 6 \tab 2538 \cr
   Total \tab 1505 \tab 93 \tab 158 \tab 288 \tab 339 \tab 125 \tab 24 \tab 6 \tab 2538 \cr
}
}

\seealso{
Other Work and Relationships: 
\code{\link{talkspvs}},
\code{\link{wkcontct}}
}
\concept{Work and Relationships}
\keyword{variable}
