% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{effctsup}
\alias{effctsup}
\title{Supervisor effective solve work/personal conflicts}
\source{
General Social Survey https://gss.norc.org
}
\description{
effctsup
}
\details{
Question How effectively does your supervisor work with employees to creatively solve conflicts between work and personal or family life…
}
\section{Values}{

\itemize{
\item \code{1} not at all effective
\item \code{2} a little
\item \code{3} somewhat
\item \code{4} very
\item \code{5} extremely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5506/vshow}{GSS Data Explorer page for \code{effctsup}}.
}

\seealso{
Other Work and Relationships: 
\code{\link{talkspvs}},
\code{\link{wkcontct}}
}
\concept{Work and Relationships}
\keyword{variable}
