% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{egypt}
\alias{egypt}
\title{Liking for egypt}
\source{
General Social Survey https://gss.norc.org
}
\description{
egypt
}
\details{
Question You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? READ EACH ITEM.
H. Egypt
}
\section{Values}{

\itemize{
\item \code{0} +5
\item \code{1} +4
\item \code{2} +3
\item \code{3} +2
\item \code{4} +1
\item \code{5} -1
\item \code{6} -2
\item \code{7} -3
\item \code{8} -4
\item \code{9} -5
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5497/vshow}{GSS Data Explorer page for \code{egypt}}.
}

\seealso{
Other War and International: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{england}},
\code{\link{israel}},
\code{\link{japan}}

Other Countries: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{england}},
\code{\link{israel}},
\code{\link{japan}}

Other Split Ballots: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{england}},
\code{\link{israel}},
\code{\link{japan}}
}
\concept{Countries}
\concept{Split Ballots}
\concept{War and International}
\keyword{variable}
