% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{eharaswk}
\alias{eharaswk}
\title{Harassed electronically at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
eharaswk
}
\details{
Question I have received emails, text messages, mobile cell phone calls, or other electronic, internet, or social network communications from people at work
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3018/vshow}{GSS Data Explorer page for \code{eharaswk}}.
}

\seealso{
Other Workplace Conflict: 
\code{\link{denyrais}},
\code{\link{ignorwk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{wkbhvrs}},
\code{\link{wkrspns}}
}
\concept{Workplace Conflict}
\keyword{variable}
