% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{eldcost}
\alias{eldcost}
\title{Who should pay for help for elderly}
\source{
General Social Survey https://gss.norc.org
}
\description{
eldcost
}
\details{
Question And who do you think should primarily cover the costs of this help to these elderly people?
}
\section{Values}{

\itemize{
\item \code{1} the elderly people themselves or their family
\item \code{2} the government/public funds
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4472/vshow}{GSS Data Explorer page for \code{eldcost}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
