% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_264.R
\name{emaillnk}
\alias{emaillnk}
\title{How often does r link in an email from friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
emaillnk
}
\details{
Question Now I'd like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit: Click on a link in an email message from a friend, relative, or co-worker.
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than half of the time
\item \code{3} about half of the time
\item \code{4} more than half of the time
\item \code{5} almost always
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2609/vshow}{GSS Data Explorer page for \code{emaillnk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than half of the time \tab about half of the time \tab more than half of the time \tab almost always \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2403 \tab 65 \tab 152 \tab 43 \tab 29 \tab 23 \tab 1 \tab 49 \tab 2765 \cr
   Total \tab 2403 \tab 65 \tab 152 \tab 43 \tab 29 \tab 23 \tab 1 \tab 49 \tab 2765 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{bookmark}},
\code{\link{catdrcty}},
\code{\link{educ30}},
\code{\link{fin30}},
\code{\link{govt30}},
\code{\link{hyperlnk}},
\code{\link{news30}},
\code{\link{schl30}},
\code{\link{work30}}
}
\concept{Information Society}
\keyword{variable}
