% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_572.R
\name{emailwrk}
\alias{emailwrk}
\title{Govt audit email and internet at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
emailwrk
}
\details{
Question Protecting secrets is a continuing concern of the government. I am going to ask you about various security measures that the government might apply to individuals with a SECRET or TOP SECRET clearance. Do you (strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree) that people with SECRET or TOP SECRET clearance should be subject to:
C. Auditing of email and Internet use at work
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5251/vshow}{GSS Data Explorer page for \code{emailwrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 276 \tab 629 \tab 209 \tab 210 \tab 32 \tab 45 \tab 18 \tab 2817 \cr
   Total \tab 1398 \tab 276 \tab 629 \tab 209 \tab 210 \tab 32 \tab 45 \tab 18 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{bugging}},
\code{\link{chkonjob}},
\code{\link{chkother}},
\code{\link{chktravl}},
\code{\link{compfin}},
\code{\link{emailhme}},
\code{\link{finanqs}},
\code{\link{lietest}},
\code{\link{testdrug}}
}
\concept{National Security}
\keyword{variable}
