% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{emailyr}
\alias{emailyr}
\title{When r first began using email}
\source{
General Social Survey https://gss.norc.org
}
\description{
emailyr
}
\details{
Question A. When did you first begin using email?
}
\section{Values}{

\itemize{
\item \code{1} this year (2002)
\item \code{2} in the last year (2001)
\item \code{3} 1-2 years ago (2000)
\item \code{4} 2-3 years ago (1999)
\item \code{5} 3-4 years ago (1998)
\item \code{6} 4-5 years ago (1997)
\item \code{7} 5 or more years ago (1990-1996)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2619/vshow}{GSS Data Explorer page for \code{emailyr}}.
}

\seealso{
Other Information Society: 
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webable}},
\code{\link{webyr}}
}
\concept{Information Society}
\keyword{variable}
