% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_184.R
\name{embarrss}
\alias{embarrss}
\title{Felt embarrassed about something?}
\source{
General Social Survey https://gss.norc.org
}
\description{
embarrss
}
\details{
Question Now I'm going to read a list of different feelings that people sometimes have. After each one, I would like you to tell me on how many days you have felt this way during the past 7 days. On how many days in the past 7 days have you... (SPECIFY NUMBER OF DAYS)
(IF RESPONDENT ASKS, SAY THAT FEELING THIS WAY AT ANY TIME DURING A DAY MAKES THAT DAY "COUNT.")
R. Felt embarrassed about something?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1723/vshow}{GSS Data Explorer page for \code{embarrss}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 986 \tab 279 \tab 108 \tab 31 \tab 9 \tab 8 \tab 6 \tab 23 \tab 3 \tab 7 \tab 2904 \cr
   Total \tab 1444 \tab 986 \tab 279 \tab 108 \tab 31 \tab 9 \tab 8 \tab 6 \tab 23 \tab 3 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angry}},
\code{\link{badbrks}},
\code{\link{litcntrl}},
\code{\link{mostluck}},
\code{\link{noemote}},
\code{\link{noplan}},
\code{\link{proud}},
\code{\link{showangr}},
\code{\link{showfeel}}
}
\concept{Emotions}
\keyword{variable}
