% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_285.R
\name{emfamloc}
\alias{emfamloc}
\title{R send or receive to or from family members at}
\source{
General Social Survey https://gss.norc.org
}
\description{
emfamloc
}
\details{
Question Thinking now about all your home, work or other computers,
G. Family members who live with you?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2601/vshow}{GSS Data Explorer page for \code{emfamloc}}.
}

\seealso{
Other Information Society: 
\code{\link{emchurch}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emgroups}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}}

Other Computers And Internet: 
\code{\link{emchurch}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emgroups}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
