% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_285.R
\name{emgroups}
\alias{emgroups}
\title{R send or receive to or from group members}
\source{
General Social Survey https://gss.norc.org
}
\description{
emgroups
}
\details{
Question Thinking now about all your home, work or other computers,
K. Members of other groups or associations to which you belong?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2605/vshow}{GSS Data Explorer page for \code{emgroups}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2423 \tab 118 \tab 217 \tab 59 \tab 2817 \cr
   2002 \tab 2167 \tab 272 \tab 285 \tab 41 \tab 2765 \cr
   Total \tab 4590 \tab 390 \tab 502 \tab 100 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emchurch}},
\code{\link{emfamloc}},
\code{\link{emfamoth}},
\code{\link{emfri}},
\code{\link{emwrkawy}},
\code{\link{emwrkloc}}
}
\concept{Information Society}
\keyword{variable}
