% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_281.R
\name{emhrh}
\alias{emhrh}
\title{Hours per week use computer email}
\source{
General Social Survey https://gss.norc.org
}
\description{
emhrh
}
\details{
Question About how many minutes or hours a week do you use your home computer for each of the following:
D. Sending or receiving e-mail?
2. Hours per week?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2563/vshow}{GSS Data Explorer page for \code{emhrh}}.

Counts by year:\tabular{llllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 15 \tab 2 \tab 20 \tab 3 \tab 35 \tab 4 \tab 40 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab 12 \tab 13 \tab 14 \tab 19 \tab 21 \tab 24 \tab 25 \tab 30 \tab 45 \tab 56 \tab 9 \tab Total \cr
   2000 \tab 2519 \tab 61 \tab 8 \tab 4 \tab 53 \tab 2 \tab 34 \tab 1 \tab 18 \tab 2 \tab 17 \tab 1 \tab 5 \tab 9 \tab 1 \tab 2 \tab 80 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1818 \tab 308 \tab 49 \tab 6 \tab 198 \tab 11 \tab 107 \tab - \tab 44 \tab 3 \tab 65 \tab 2 \tab 21 \tab 52 \tab 12 \tab 3 \tab 44 \tab 4 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2765 \cr
   Total \tab 4337 \tab 369 \tab 57 \tab 10 \tab 251 \tab 13 \tab 141 \tab 1 \tab 62 \tab 5 \tab 82 \tab 3 \tab 26 \tab 61 \tab 13 \tab 5 \tab 124 \tab 4 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compwork}},
\code{\link{emminh}},
\code{\link{jobhrw}},
\code{\link{jobminw}},
\code{\link{perhrh}},
\code{\link{wrkhrh}},
\code{\link{wrkminh}},
\code{\link{wwwhrh}},
\code{\link{wwwminh}}
}
\concept{Information Society}
\keyword{variable}
