% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_448.R
\name{emonitor}
\alias{emonitor}
\title{Civil liberties - monitor emails}
\source{
General Social Survey https://gss.norc.org
}
\description{
emonitor
}
\details{
Question (Do you think that the American government should or should not have the right to do the following?)
Monitor e-mails and any other information exchanged on the Internet?  (Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} definitely should have the right
\item \code{2} probably should have the right
\item \code{3} probably should not have the right
\item \code{4} definitely should not have the right
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6316/vshow}{GSS Data Explorer page for \code{emonitor}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should have the right \tab probably should have the right \tab probably should not have the right \tab definitely should not have the right \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 125 \tab 315 \tab 410 \tab 511 \tab 16 \tab 13 \tab 2867 \cr
   Total \tab 1477 \tab 125 \tab 315 \tab 410 \tab 511 \tab 16 \tab 13 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cctv}},
\code{\link{cutdebt}},
\code{\link{govtinfo}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
