% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{emonitor1}
\alias{emonitor1}
\title{Should government be able to monitor online communication}
\source{
General Social Survey https://gss.norc.org
}
\description{
emonitor1
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} definitely should have the right
\item \code{2} probably should have the right
\item \code{3} probably should not have the right
\item \code{4} definitely should not have the right
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
