% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{emotfeel}
\alias{emotfeel}
\title{Treatment for emotion improve feeling about}
\source{
General Social Survey https://gss.norc.org
}
\description{
emotfeel
}
\details{
Question Say you went to treatment for an EMOTIONAL HEALTH problem such as feeling depressed or anxious that affects your work and other daily activities so that you accomplish less than you would like...
E. Would you definitely expect, probably expect, probably not expect, or definitely not expect to feel better about yourself as a person?
}
\section{Values}{

\itemize{
\item \code{1} definitely expect
\item \code{2} probably expect
\item \code{3} probably not expect
\item \code{4} definitely not expect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2724/vshow}{GSS Data Explorer page for \code{emotfeel}}.
}

\seealso{
Other Health Status: 
\code{\link{emotrely}}

Other Health: 
\code{\link{emotrely}}
}
\concept{Health}
\concept{Health Status}
\keyword{variable}
