% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{empathy1}
\alias{empathy1}
\title{R have tender feeling for people less fortunate}
\source{
General Social Survey https://gss.norc.org
}
\description{
empathy1
}
\details{
Question The following statements ask about your thoughts and feelings in various situations. For each item indicate how well it describes you by choosing the number on the showcard where 1 indicates that it does not describe you very well  and 5 means that it does describe you very well. Of course numbers 2-4 indicate that how well it describes you are In between these points.
A. I often have tender, concerned feelings for people less fortunate than me.
}
\section{Values}{

\itemize{
\item \verb{[1]} does not describe very well
\item \verb{[5]} describes very well
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2870/vshow}{GSS Data Explorer page for \code{empathy1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab does not describe very well \tab 2 \tab 3 \tab 4 \tab describes very well \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 64 \tab 68 \tab 253 \tab 341 \tab 627 \tab 13 \tab 6 \tab 2765 \cr
   2004 \tab 1472 \tab 44 \tab 59 \tab 201 \tab 418 \tab 609 \tab 5 \tab 4 \tab 2812 \cr
   Total \tab 2865 \tab 108 \tab 127 \tab 454 \tab 759 \tab 1236 \tab 18 \tab 10 \tab 5577 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{empathy2}},
\code{\link{empathy3}},
\code{\link{empathy4}}
}
\concept{Altruism}
\keyword{variable}
