% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_323.R
\name{empathy5}
\alias{empathy5}
\title{R does not feel pity for someone treated unfair}
\source{
General Social Survey https://gss.norc.org
}
\description{
empathy5
}
\details{
Question The following statements ask about your thoughts and feelings in various situations. For each item indicate how well it describes you by choosing the number on the showcard where 1 indicates that it does not describe you very well  and 5 means that it does describe you very well. Of course numbers 2-4 indicate that how well it describes you are In between these points.
E. When I see someone being treated unfairly, I sometimes don't feel very much pity for them.
}
\section{Values}{

\itemize{
\item \code{1} does not describe very well
\item \code{5} describes very well
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2874/vshow}{GSS Data Explorer page for \code{empathy5}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab does not describe very well \tab 2 \tab 3 \tab 4 \tab describes very well \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 613 \tab 373 \tab 191 \tab 90 \tab 76 \tab 21 \tab 8 \tab 2765 \cr
   2004 \tab 1472 \tab 633 \tab 374 \tab 163 \tab 89 \tab 68 \tab 8 \tab 5 \tab 2812 \cr
   Total \tab 2865 \tab 1246 \tab 747 \tab 354 \tab 179 \tab 144 \tab 29 \tab 13 \tab 5577 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{cutahead}},
\code{\link{empathy6}},
\code{\link{empathy7}},
\code{\link{givblood}},
\code{\link{givchrty}},
\code{\link{givhmlss}},
\code{\link{givseat}},
\code{\link{retchnge}},
\code{\link{volchrty}}
}
\concept{Altruism}
\keyword{variable}
