% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{emphlth}
\alias{emphlth}
\title{Do you receive health insurance from employer}
\source{
General Social Survey https://gss.norc.org
}
\description{
emphlth
}
\details{
Question Do you receive health insurance from your employer?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3903/vshow}{GSS Data Explorer page for \code{emphlth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 949 \tab 652 \tab 371 \tab 3 \tab 48 \tab 2023 \cr
   Total \tab 949 \tab 652 \tab 371 \tab 3 \tab 48 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphplan}},
\code{\link{natdeg}},
\code{\link{othplan}},
\code{\link{typevisa}},
\code{\link{visa}}
}
\concept{Global Economics}
\keyword{variable}
