% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{emphlth2}
\alias{emphlth2}
\title{Employers coverage}
\source{
General Social Survey https://gss.norc.org
}
\description{
emphlth2
}
\details{
Question Which of the following statements best describes your employer’s coverage:
}
\section{Values}{

\itemize{
\item \code{1} only individual employees are covered
\item \code{2} employees and their spouses are covered but domestic partners are not covered
\item \code{3} employees and their spouses are covered and domestic partners are covered
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3929/vshow}{GSS Data Explorer page for \code{emphlth2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab only individual employees are covered \tab employees and their spouses are covered but domestic partners are not covered \tab employees and their spouses are covered and domestic partners are covered \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1945 \tab 14 \tab 27 \tab 8 \tab 3 \tab 26 \tab 2023 \cr
   Total \tab 1945 \tab 14 \tab 27 \tab 8 \tab 3 \tab 26 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{emphlth1}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{mykids}},
\code{\link{othkids}},
\code{\link{sppart}}
}
\concept{Sexual Orientation}
\keyword{variable}
