% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{empinput}
\alias{empinput}
\title{R involved in any task force for decision-making}
\source{
General Social Survey https://gss.norc.org
}
\description{
empinput
}
\details{
Question Some companies have organized workplace decision-making in ways to get more employee input and involvement. Are you personally involved in any group,  team, committee, or task force that  addresses issues such as product quality, cost cutting, productivity, health and safety, or other workplace issues?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2014, 2018 \cr
   B/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3557/vshow}{GSS Data Explorer page for \code{empinput}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2006 \tab 2778 \tab 541 \tab 1143 \tab 27 \tab 21 \tab - \tab 4510 \cr
   2014 \tab 1765 \tab 227 \tab 541 \tab 2 \tab 3 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 361 \tab 1054 \tab 3 \tab 30 \tab - \tab 2348 \cr
   2022 \tab 2570 \tab 254 \tab 657 \tab 46 \tab 9 \tab 8 \tab 3544 \cr
   Total \tab 8013 \tab 1383 \tab 3395 \tab 78 \tab 63 \tab 8 \tab 12940 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
