% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{emptrain}
\alias{emptrain}
\title{Received formal training from employer}
\source{
General Social Survey https://gss.norc.org
}
\description{
emptrain
}
\details{
Question B. In the last 12 months have you received any formal training from your current employer, such as in classes or seminars sponsored by the employer?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2014, 2018 \cr
   B/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3559/vshow}{GSS Data Explorer page for \code{emptrain}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2006 \tab 2778 \tab 832 \tab 864 \tab 19 \tab 17 \tab - \tab 4510 \cr
   2014 \tab 1765 \tab 381 \tab 389 \tab 1 \tab 2 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 739 \tab 677 \tab 2 \tab 30 \tab - \tab 2348 \cr
   2022 \tab 2570 \tab 453 \tab 486 \tab 23 \tab 6 \tab 6 \tab 3544 \cr
   Total \tab 8013 \tab 2405 \tab 2416 \tab 45 \tab 55 \tab 6 \tab 12940 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
