% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{endsme12}
\alias{endsme12}
\title{During the next 12 months, how difficult r's household to make ends meet}
\source{
General Social Survey https://gss.norc.org
}
\description{
endsme12
}
\details{
Question During the next 12 months, how difficult or easy do you think it will be for your household to make ends meet?
}
\section{Values}{

\itemize{
\item \verb{[1]} very difficult
\item \verb{[2]} fairly difficult
\item \verb{[3]} neither easy nor difficult
\item \verb{[4]} fairly easy
\item \verb{[5]} very easy
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7518/vshow}{GSS Data Explorer page for \code{endsme12}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very difficult \tab fairly difficult \tab neither easy nor difficult \tab fairly easy \tab very easy \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 132 \tab 349 \tab 507 \tab 488 \tab 276 \tab 59 \tab 176 \tab 16 \tab 4032 \cr
   Total \tab 2029 \tab 132 \tab 349 \tab 507 \tab 488 \tab 276 \tab 59 \tab 176 \tab 16 \tab 4032 \cr
}
}

\keyword{variable}
