% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{endsmeet}
\alias{endsmeet}
\title{How difficult or easy is it for r's household to make ends meet}
\source{
General Social Survey https://gss.norc.org
}
\description{
endsmeet
}
\details{
Question Thinking of your household's total income, including all the sources of income of all the members who contribute to it, how difficult or easy is it currently for your household to make ends meet?
}
\section{Values}{

\itemize{
\item \code{1} very difficult
\item \code{2} fairly difficult
\item \code{3} neither easy nor difficult
\item \code{4} fairly easy
\item \code{5} very easy
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7276/vshow}{GSS Data Explorer page for \code{endsmeet}}.
}

\keyword{variable}
