% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{engballt}
\alias{engballt}
\title{Attitude toward printing ballots only in english}
\source{
General Social Survey https://gss.norc.org
}
\description{
engballt
}
\details{
Question Do you believe that election ballots should be printed only in English, or in areas where lots of people don't speak English, should they also be printed in some other languages?
}
\section{Values}{

\itemize{
\item \verb{[1]} only in english
\item \verb{[2]} in other langs
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1609/vshow}{GSS Data Explorer page for \code{engballt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab only in english \tab in other langs \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 524 \tab 898 \tab 43 \tab 9 \tab 2992 \cr
   Total \tab 1518 \tab 524 \tab 898 \tab 43 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
