% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{engballt}
\alias{engballt}
\title{Attitude toward printing ballots only in english}
\source{
General Social Survey https://gss.norc.org
}
\description{
engballt
}
\details{
Question Do you believe that election ballots should be printed only in English, or in areas where lots of people don't speak English, should they also be printed in some other languages?
}
\section{Values}{

\itemize{
\item \code{1} only in english
\item \code{2} in other langs
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1609/vshow}{GSS Data Explorer page for \code{engballt}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
