% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_389.R
\name{engbrng}
\alias{engbrng}
\title{Being engineer boring}
\source{
General Social Survey https://gss.norc.org
}
\description{
engbrng
}
\details{
Question K. A job as an engineer would be boring. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3536/vshow}{GSS Data Explorer page for \code{engbrng}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 12 \tab 49 \tab 331 \tab 36 \tab 54 \tab 3 \tab 1974 \cr
   Total \tab 1489 \tab 12 \tab 49 \tab 331 \tab 36 \tab 54 \tab 3 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engbtr}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{engint}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{farming}},
\code{\link{journlsm}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
