% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_388.R
\name{engdo}
\alias{engdo}
\title{Know what engineers do}
\source{
General Social Survey https://gss.norc.org
}
\description{
engdo
}
\details{
Question D. Next, I’d like to ask how you rate your knowledge of what engineers do day-to-day in their jobs. would you say your knowledge of what engineers do day-to-day on their jobs is Excellent, Good, Fair, Poor, Very Poor?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} good
\item \code{3} fair
\item \code{4} poor
\item \code{5} very poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3525/vshow}{GSS Data Explorer page for \code{engdo}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}

Other Employment: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}

Other Job: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}

Other Occupation: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}

Other Work: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}
}
\concept{Employment}
\concept{Job}
\concept{Occupation}
\concept{Science Knowledge & Attitudes}
\concept{Work}
\keyword{variable}
