% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_254.R
\name{engoff1}
\alias{engoff1}
\title{English official language of u.s.}
\source{
General Social Survey https://gss.norc.org
}
\description{
engoff1
}
\details{
Question Do you favor a law making English the official language of the United States, or do you oppose such a law?
}
\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2294/vshow}{GSS Data Explorer page for \code{engoff1}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{forlang1}},
\code{\link{granlang}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{nobiling}},
\code{\link{twolang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
