% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{engoffcl}
\alias{engoffcl}
\title{Favor law making english the official language}
\source{
General Social Survey https://gss.norc.org
}
\description{
engoffcl
}
\details{
Question Do you favor a law making English the official language of the United States, meaning government business would be conducted in English only, or do you oppose such a law?
}
\section{Values}{

\itemize{
\item \verb{[1]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[5]} oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1610/vshow}{GSS Data Explorer page for \code{engoffcl}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab favor \tab neither favor nor oppose \tab oppose \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 890 \tab 137 \tab 391 \tab 50 \tab 6 \tab 2992 \cr
   Total \tab 1518 \tab 890 \tab 137 \tab 391 \tab 50 \tab 6 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
