% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_388.R
\name{engresp}
\alias{engresp}
\title{Consider work in engineer field}
\source{
General Social Survey https://gss.norc.org
}
\description{
engresp
}
\details{
Question C. Have you ever considered working in an engineering-related career?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3524/vshow}{GSS Data Explorer page for \code{engresp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab can't choose \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 126 \tab 355 \tab 1 \tab 3 \tab 1974 \cr
   Total \tab 1489 \tab 126 \tab 355 \tab 1 \tab 3 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engson}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
