% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_388.R
\name{engson}
\alias{engson}
\title{Happy if son engineer}
\source{
General Social Survey https://gss.norc.org
}
\description{
engson
}
\details{
Question B. If you had a son, how would you feel if he wanted to be an engineer—would you feel happy, unhappy, or would you not care one way or the other?
}
\section{Values}{

\itemize{
\item \verb{[1]} happy
\item \verb{[2]} not care
\item \verb{[3]} unhappy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3523/vshow}{GSS Data Explorer page for \code{engson}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab happy \tab not care \tab unhappy \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 419 \tab 61 \tab 1 \tab 1 \tab 3 \tab 1974 \cr
   Total \tab 1489 \tab 419 \tab 61 \tab 1 \tab 1 \tab 3 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{enggood}},
\code{\link{englone}},
\code{\link{engresp}},
\code{\link{scientbr}},
\code{\link{scientmo}},
\code{\link{scientwk}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
