% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{engteach}
\alias{engteach}
\title{Opinion on teaching classes only in english}
\source{
General Social Survey https://gss.norc.org
}
\description{
engteach
}
\details{
Question There are several different ideas about how to teach children who don't speak English when they enter our public schools. Please look at the card and tell me which statement best describes how you feel:
All classes should be conducted only in English so that children have to learn English right from the start…………………………………………………….…………………1
Children who don't know English should have classes in their native language just for a year or two until they learn English………………...……….….....…………… 2
Students who want to keep up with their native languages and cultures should be able to take many of the classes in Spanish or other languages All the way through high school……......3
}
\section{Values}{

\itemize{
\item \code{1} all classes in english
\item \code{2} 1-2 yrs in native lang
\item \code{3} native lang thru hs
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1608/vshow}{GSS Data Explorer page for \code{engteach}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab all classes in english \tab 1-2 yrs in native lang \tab native lang thru hs \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 512 \tab 687 \tab 223 \tab 41 \tab 11 \tab 2992 \cr
   Total \tab 1518 \tab 512 \tab 687 \tab 223 \tab 41 \tab 11 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
