% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{enjoynat}
\alias{enjoynat}
\title{How much do you enjoy being outside in nature?}
\source{
General Social Survey https://gss.norc.org
}
\description{
enjoynat
}
\details{
Question How much, if at all, do you enjoy being outside in nature?
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} to a small extent
\item \code{3} to some extent
\item \code{4} to a great extent
\item \code{5} to a very great extent
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7495/vshow}{GSS Data Explorer page for \code{enjoynat}}.
}

\keyword{variable}
