% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{enprbfam}
\alias{enprbfam}
\title{Problem effect family most}
\source{
General Social Survey https://gss.norc.org
}
\description{
enprbfam
}
\details{
Question E. Which problem, if any, affects you and your family the most?
}
\section{Values}{

\itemize{
\item \verb{[1]} air pollution
\item \verb{[2]} chemicals and pesticides
\item \verb{[3]} water shortage
\item \verb{[4]} water pollution
\item \verb{[5]} nuclear waste
\item \verb{[6]} domestic waste disposal
\item \verb{[7]} climate change
\item \verb{[8]} genetically modified foods
\item \verb{[9]} using up our natural resources
\item \verb{[10]} none of these
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4804/vshow}{GSS Data Explorer page for \code{enprbfam}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab air pollution \tab chemicals and pesticides \tab water shortage \tab water pollution \tab nuclear waste \tab domestic waste disposal \tab climate change \tab genetically modified foods \tab using up our natural resources \tab none of these \tab don't know \tab no answer \tab Total \cr
   2010 \tab 614 \tab 300 \tab 159 \tab 71 \tab 124 \tab 15 \tab 65 \tab 85 \tab 132 \tab 150 \tab 123 \tab 199 \tab 7 \tab 2044 \cr
   Total \tab 614 \tab 300 \tab 159 \tab 71 \tab 124 \tab 15 \tab 65 \tab 85 \tab 132 \tab 150 \tab 123 \tab 199 \tab 7 \tab 2044 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
