% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{enprbus}
\alias{enprbus}
\title{Most important environment problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
enprbus
}
\details{
Question D. Here is a list of some different environmental problems. Which problem, if any, do you think is the most important for America as a whole?
}
\section{Values}{

\itemize{
\item \verb{[1]} air pollution
\item \verb{[2]} chemicals and pesticides
\item \verb{[3]} water shortage
\item \verb{[4]} water pollution
\item \verb{[5]} nuclear waste
\item \verb{[6]} domestic waste disposal
\item \verb{[7]} climate change
\item \verb{[8]} genetically modified foods
\item \verb{[9]} using up our natural resources
\item \verb{[10]} none of these
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4803/vshow}{GSS Data Explorer page for \code{enprbus}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab air pollution \tab chemicals and pesticides \tab water shortage \tab water pollution \tab nuclear waste \tab domestic waste disposal \tab climate change \tab genetically modified foods \tab using up our natural resources \tab none of these \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 199 \tab 112 \tab 95 \tab 134 \tab 88 \tab 60 \tab 127 \tab 61 \tab 344 \tab 22 \tab 181 \tab 7 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 141 \tab 140 \tab 76 \tab 131 \tab 40 \tab 87 \tab 642 \tab 115 \tab 177 \tab 93 \tab 217 \tab 160 \tab 10 \tab 4032 \cr
   Total \tab 2617 \tab 340 \tab 252 \tab 171 \tab 265 \tab 128 \tab 147 \tab 769 \tab 176 \tab 521 \tab 115 \tab 398 \tab 167 \tab 10 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
