% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_342.R
\name{entity}
\alias{entity}
\title{Entity mentions in religious change}
\source{
General Social Survey https://gss.norc.org
}
\description{
entity
}
\details{
Question Open-ended questions for how people who had undergone spiritual or religious change described their experiences. Specific religious figures mentioned in religious change.
}
\section{Values}{

\itemize{
\item \code{96} no mention
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3067/vshow}{GSS Data Explorer page for \code{entity}}.
}

\seealso{
Other Spirituality and Religion: 
\code{\link{agerborn}},
\code{\link{chngrbrn}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{numrborn}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}}
}
\concept{Spirituality and Religion}
\keyword{variable}
