% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{eqclose}
\alias{eqclose}
\title{R feels equally close to all persons}
\source{
General Social Survey https://gss.norc.org
}
\description{
eqclose
}
\details{
Question Do you feel equally close to all these people?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[7]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/849/vshow}{GSS Data Explorer page for \code{eqclose}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1985 \tab 364 \tab 639 \tab 525 \tab 6 \tab 1534 \cr
   1987 \tab 385 \tab 824 \tab 540 \tab 70 \tab 1819 \cr
   Total \tab 749 \tab 1463 \tab 1065 \tab 76 \tab 3353 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{numgiven}}
}
\concept{Social Networks}
\keyword{variable}
