% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{eqincome}
\alias{eqincome}
\title{Govt should redistribute wealth}
\source{
General Social Survey https://gss.norc.org
}
\description{
eqincome
}
\details{
Question What is your opinion of the following statement?
It is the responsibility of the government to reduce the differences in income between people with high incomes and  those with low incomes.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4005/vshow}{GSS Data Explorer page for \code{eqincome}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{databank}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{progtax}},
\code{\link{verdict}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
