% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_28.R
\name{equal6}
\alias{equal6}
\title{One is able to live well in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
equal6
}
\details{
Question SHUFFLE THE GREY CARDS. READ STATEMENTS IN ORDER OF CARDS TO RESPONDENT. CIRCLE APPROPRIATE CODE FOR EACH STATEMENT.
F. All in all, one can live well in America.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/230/vshow}{GSS Data Explorer page for \code{equal6}}.
}

\seealso{
Other Core: 
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{natmassz}},
\code{\link{natparkz}},
\code{\link{natroadz}},
\code{\link{natsocz}}

Other Corporations: 
\code{\link{equal2}}

Other Social Class: 
\code{\link{equal1}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}}
}
\concept{Core}
\concept{Corporations}
\concept{Social Class}
\concept{Social Mobility}
\keyword{variable}
