% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_444.R
\name{equalize}
\alias{equalize}
\title{Govts resp: reduce income differences}
\source{
General Social Survey https://gss.norc.org
}
\description{
equalize
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to . . .
G. Reduce income differences between the rich and poor.
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1991, 1996, 1998 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4082/vshow}{GSS Data Explorer page for \code{equalize}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 106 \tab 126 \tab 160 \tab 210 \tab 60 \tab 15 \tab 1534 \cr
   1990 \tab 155 \tab 186 \tab 304 \tab 330 \tab 283 \tab 72 \tab 42 \tab 1372 \cr
   1991 \tab 158 \tab 262 \tab 329 \tab 319 \tab 317 \tab 99 \tab 33 \tab 1517 \cr
   1996 \tab 1572 \tab 206 \tab 363 \tab 314 \tab 304 \tab 85 \tab 60 \tab 2904 \cr
   1998 \tab 1548 \tab 216 \tab 332 \tab 299 \tab 321 \tab 97 \tab 19 \tab 2832 \cr
   2006 \tab 2992 \tab 423 \tab 349 \tab 394 \tab 311 \tab 40 \tab 1 \tab 4510 \cr
   2016 \tab 1477 \tab 366 \tab 382 \tab 338 \tab 252 \tab 38 \tab 14 \tab 2867 \cr
   Total \tab 8759 \tab 1765 \tab 2185 \tab 2154 \tab 1998 \tab 491 \tab 184 \tab 17536 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{hlthcare}},
\code{\link{protstrs}},
\code{\link{racists}},
\code{\link{revoltrs}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
