% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{eqwlth}
\alias{eqwlth}
\title{Should govt reduce income differences}
\source{
General Social Survey https://gss.norc.org
}
\description{
eqwlth
}
\details{
Question A. Some people think that the government in Washington ought to reduce the income differences between the rich and the poor, perhaps by raising the taxes of wealthy families or by giving income assistance to the poor. Others think that the government should not concern itself with reducing this income difference between the rich and the poor. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that the government ought to reduce the income differences between rich and poor, and a score of 7 meaning that the government should not concern itself with reducing income differences. What score between 1 and 7 comes closest to the way you feel? (CIRCLE ONE):
}
\section{Values}{

\itemize{
\item \code{1} the government should reduce income differences
\item \code{7} the government should not concern itself with reducing income differences
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
   1996 \tab B/C \tab partial \cr
   1998 \tab B/C \tab partial \cr
   2000 \tab B/C \tab partial \cr
   2002 \tab B/C \tab partial \cr
   2004 \tab B/C \tab partial \cr
   2006 \tab B/C \tab partial \cr
   2008 \tab B/C \tab partial \cr
   2010 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab B/C \tab partial \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/243/vshow}{GSS Data Explorer page for \code{eqwlth}}.
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{educop}},
\code{\link{eqwlthy}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}},
\code{\link{tax}}
}
\concept{Core}
\keyword{variable}
