% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{era}
\alias{era}
\title{Does r favor or oppose era?}
\source{
General Social Survey https://gss.norc.org
}
\description{
era
}
\details{
Question Have you heard or read about the Equal Rights Amendment?
B. Do you strongly favor, somewhat favor, somewhat oppose, or strongly oppose this amendment?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} oppose
\item \verb{[4]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/709/vshow}{GSS Data Explorer page for \code{era}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1977 \tab 209 \tab 271 \tab 606 \tab 201 \tab 118 \tab 110 \tab 15 \tab 1530 \cr
   1982 \tab 183 \tab 438 \tab 730 \tab 267 \tab 134 \tab 84 \tab 24 \tab 1860 \cr
   Total \tab 392 \tab 709 \tab 1336 \tab 468 \tab 252 \tab 194 \tab 39 \tab 3390 \cr
}
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{ablegal}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{eratell}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}
}
\concept{Core}
\keyword{variable}
