% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{eraimp}
\alias{eraimp}
\title{Importance of era issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
eraimp
}
\details{
Question How important is the Equal Rights Amendment issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \code{1} one of most imp
\item \code{2} important
\item \code{3} not very imp
\item \code{4} not important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/714/vshow}{GSS Data Explorer page for \code{eraimp}}.
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubgv}},
\code{\link{racsubs}}
}
\concept{Core}
\keyword{variable}
