% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{erainfo}
\alias{erainfo}
\title{How much info does r have about era}
\source{
General Social Survey https://gss.norc.org
}
\description{
erainfo
}
\details{
Question How much information do you have about the Equal Rights Amendment? Do you have all the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \code{1} all info r needs
\item \code{2} most of the info
\item \code{3} some info
\item \code{4} very little info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/715/vshow}{GSS Data Explorer page for \code{erainfo}}.
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubgv}},
\code{\link{racsubs}}

Other Equal Rights Amendment: 
\code{\link{eraimp}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}}

Other Split Ballots: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}}
}
\concept{Core}
\concept{Equal Rights Amendment}
\concept{Split Ballots}
\keyword{variable}
