% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{eratell}
\alias{eratell}
\title{Does r favor or oppose era}
\source{
General Social Survey https://gss.norc.org
}
\description{
eratell
}
\details{
Question Have you heard or read about the Equal Rights Amendment?
C. The Equal Rights Amendment will amend the United States Constitution to provide equal rights for men and women. Do you strongly favor, somewhat favor, somewhat oppose, or strongly oppose this amendment?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} somewhat favor
\item \code{3} somewht oppose
\item \code{4} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/710/vshow}{GSS Data Explorer page for \code{eratell}}.
}

\seealso{
Other Core: 
\code{\link{abhave3}},
\code{\link{ablegal}},
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}},
\code{\link{fechld}},
\code{\link{fefam}},
\code{\link{fehelp}},
\code{\link{fepresch}}

Other Equal Rights Amendment: 
\code{\link{era}},
\code{\link{erameans}},
\code{\link{eraread}}
}
\concept{Core}
\concept{Equal Rights Amendment}
\keyword{variable}
