% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{erosion}
\alias{erosion}
\title{Example of erosion}
\source{
General Social Survey https://gss.norc.org
}
\description{
erosion
}
\details{
Question Example of Erosion.
}
\section{Values}{

\itemize{
\item \code{1} the wind in the desert blows sand against a rock
\item \code{2} a glacier picks up boulders as it moves
\item \code{3} a flood washes over a riverbank, and the water carries small soil particles downstream
\item \code{4} an icy winter causes the pavement in a road to crack
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3478/vshow}{GSS Data Explorer page for \code{erosion}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
