% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{esp}
\alias{esp}
\title{Felt in touch with someone far away}
\source{
General Social Survey https://gss.norc.org
}
\description{
esp
}
\details{
Question How often have you had any of the following experiences?  READ EACH ITEM AND CIRCLE ONE CODE FOR EACH.
B. Felt as though you were in touch with someone when they were far away  from you.
}
\section{Values}{

\itemize{
\item \code{1} never in my life
\item \code{2} once or twice
\item \code{3} several times
\item \code{4} often
\item \code{8} cannot answer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/317/vshow}{GSS Data Explorer page for \code{esp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never in my life \tab once or twice \tab several times \tab often \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 480 \tab 411 \tab 412 \tab 136 \tab 15 \tab 19 \tab 1473 \cr
   1988 \tab - \tab 512 \tab 515 \tab 313 \tab 116 \tab 20 \tab 5 \tab 1481 \cr
   1989 \tab 531 \tab 415 \tab 299 \tab 212 \tab 66 \tab 11 \tab 3 \tab 1537 \cr
   Total \tab 531 \tab 1407 \tab 1225 \tab 937 \tab 318 \tab 46 \tab 27 \tab 4491 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
