% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{ethadapt}
\alias{ethadapt}
\title{Minorities must adapt to american culture}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethadapt
}
\details{
Question Here are some opinions some people have expressed in connection with ethnic issues in the United States. To what extent do you agree or disagree with each one?
D. In order to have a smoothly functioning society, members of ethnic minorities must better adapt to the ways of mainstream American culture.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2729/vshow}{GSS Data Explorer page for \code{ethadapt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 412 \tab 1210 \tab 546 \tab 446 \tab 99 \tab 42 \tab 10 \tab 2765 \cr
   Total \tab 0 \tab 412 \tab 1210 \tab 546 \tab 446 \tab 99 \tab 42 \tab 10 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethnofit}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelblks}},
\code{\link{feelhsps}}
}
\concept{Prejudice}
\keyword{variable}
