% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_257.R
\name{ethchng}
\alias{ethchng}
\title{Ethnic change in 25 yrs good for country}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethchng
}
\details{
Question When you think about these changes in the racial and ethnic make-up of the country in the next 25 years, do you think they will be a very good thing for the country, a good thing, neither good nor bad, a bad thing, or a very bad thing?
}
\section{Values}{

\itemize{
\item \code{1} very good thing
\item \code{2} good thing
\item \code{3} neither good nor bad
\item \code{4} bad thing
\item \code{5} very bad thing
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2322/vshow}{GSS Data Explorer page for \code{ethchng}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{betrlang}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blkwrk}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{knwwht}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{leastcom}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{nobiling}},
\code{\link{othbest}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{uselang}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{uswht}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
