% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{ethclose}
\alias{ethclose}
\title{How close feel to your ethnic or racial group?}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethclose
}
\details{
Question How close do you feel to your ethnic or racial group?
}
\section{Values}{

\itemize{
\item \code{1} very close
\item \code{2} close
\item \code{3} not very close
\item \code{4} not close at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4875/vshow}{GSS Data Explorer page for \code{ethclose}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{parcit}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}}
}
\concept{Ethnicity}
\concept{ISSP National Identity}
\keyword{variable}
