% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{ethdiff}
\alias{ethdiff}
\title{Ethnic groups in us different}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethdiff
}
\details{
Question Think about the various ethnic groups within the U.S. including Whites, African Americans, Asian Americans, Hispanics, Native Americans, and so on. To what extent do you agree with the following statement.
B. Now think only about ethnic minority groups within the U.S. such as African Americans, Asian Americans, Hispanics, and Native Americans. To what extent do you agree with the following statement. Ethnic minority groups in the U.S. are very distinct and very different from one another.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2735/vshow}{GSS Data Explorer page for \code{ethdiff}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 278 \tab 1378 \tab 575 \tab 445 \tab 37 \tab 44 \tab 8 \tab 2765 \cr
   Total \tab 0 \tab 278 \tab 1378 \tab 575 \tab 445 \tab 37 \tab 44 \tab 8 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethsame}},
\code{\link{feelwhts}},
\code{\link{whtsdiff}}
}
\concept{Prejudice}
\keyword{variable}
