% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{ethdiff}
\alias{ethdiff}
\title{Ethnic groups in us different}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethdiff
}
\details{
Question Think about the various ethnic groups within the U.S. including Whites, African Americans, Asian Americans, Hispanics, Native Americans, and so on. To what extent do you agree with the following statement.
B. Now think only about ethnic minority groups within the U.S. such as African Americans, Asian Americans, Hispanics, and Native Americans. To what extent do you agree with the following statement. Ethnic minority groups in the U.S. are very distinct and very different from one another.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2735/vshow}{GSS Data Explorer page for \code{ethdiff}}.
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethnofit}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelblks}},
\code{\link{feelhsps}},
\code{\link{feelwhts}},
\code{\link{whtsdiff}}
}
\concept{Prejudice}
\keyword{variable}
