% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{ethhist}
\alias{ethhist}
\title{Minority history adequately taught in school}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethhist
}
\details{
Question In history classes in high school and college, do the experiences of racial and ethnic minority groups in America receive too much attention now, too little attention, or about the right amount?
}
\section{Values}{

\itemize{
\item \code{1} too much attention now
\item \code{2} too little attention
\item \code{3} about the right amount
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1630/vshow}{GSS Data Explorer page for \code{ethhist}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
