% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_170.R
\name{ethid}
\alias{ethid}
\title{R american vs ethnic identity}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethid
}
\details{
Question When you think of social and political issues, do you think of yourself mainly as a member of a particular ethnic, racial, or nationality group or do you think of yourself mainly as just an American?
}
\section{Values}{

\itemize{
\item \verb{[1]} just an american
\item \verb{[2]} some part. ethnic grp
\item \verb{[3]} 2 or more ethnic grps
\item \verb{[4]} amer and ethnic grps
\item \verb{[5]} it depends
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1591/vshow}{GSS Data Explorer page for \code{ethid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab just an american \tab some part. ethnic grp \tab 2 or more ethnic grps \tab amer and ethnic grps \tab it depends \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 1315 \tab 109 \tab 3 \tab 20 \tab 12 \tab 8 \tab 7 \tab 2992 \cr
   Total \tab 1518 \tab 1315 \tab 109 \tab 3 \tab 20 \tab 12 \tab 8 \tab 7 \tab 2992 \cr
}
}

\concept{Multiculturalism}
\keyword{variable}
