% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{ethid1}
\alias{ethid1}
\title{R primary identity as ethnic or american}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethid1
}
\details{
Question C. Well, if you had to choose, which would you think of yourself most of the time--mainly as a member of a particular ethnic, racial, or nationality group, or mainly as just an American?
}
\section{Values}{

\itemize{
\item \code{1} just an american
\item \code{2} some part. ethnic grp
\item \code{3} 2 or more ethnic grps
\item \code{4} it depends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1594/vshow}{GSS Data Explorer page for \code{ethid1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab just an american \tab some part. ethnic grp \tab it depends \tab no answer \tab Total \cr
   1994 \tab 2953 \tab 13 \tab 11 \tab 6 \tab 9 \tab 2992 \cr
   Total \tab 2953 \tab 13 \tab 11 \tab 6 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}},
\code{\link{meltpot}}
}
\concept{Multiculturalism}
\keyword{variable}
