% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{ethimp}
\alias{ethimp}
\title{Importance of ethnic group to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethimp
}
\details{
Question When you think about yourself, how important is your ethnic group membership to your sense of who you are?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} moderately important
\item \code{3} slightly important
\item \code{4} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2725/vshow}{GSS Data Explorer page for \code{ethimp}}.
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethdiff}},
\code{\link{ethignor}},
\code{\link{ethnofit}},
\code{\link{ethsame}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelblks}},
\code{\link{feelhsps}},
\code{\link{feelwhts}},
\code{\link{whtsdiff}}
}
\concept{Prejudice}
\keyword{variable}
