% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{ethnic}
\alias{ethnic}
\title{Country of family origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethnic
}
\details{
Question From what countries or part of the world did your ancestors come?
A. IF MORE THAN ONE COUNTRY NAMED:Which one of these countries do you feel closer to?
}
\section{Values}{

\itemize{
\item \verb{[1]} africa (general)
\item \verb{[2]} austria
\item \verb{[3]} french canada
\item \verb{[4]} other canada
\item \verb{[5]} china
\item \verb{[6]} czechoslovakia
\item \verb{[7]} denmark
\item \verb{[8]} england & wales
\item \verb{[9]} finland
\item \verb{[10]} france
\item \verb{[11]} germany
\item \verb{[12]} greece
\item \verb{[13]} hungary
\item \verb{[14]} ireland
\item \verb{[15]} italy
\item \verb{[16]} japan
\item \verb{[17]} mexico
\item \verb{[18]} netherlands
\item \verb{[19]} norway
\item \verb{[20]} philippines
\item \verb{[21]} poland
\item \verb{[22]} puerto rico
\item \verb{[23]} russia
\item \verb{[24]} scotland
\item \verb{[25]} spain
\item \verb{[26]} sweden
\item \verb{[27]} switzerland
\item \verb{[28]} west indies
\item \verb{[29]} other
\item \verb{[30]} american indian
\item \verb{[31]} india
\item \verb{[32]} portugal
\item \verb{[33]} lithuania
\item \verb{[34]} yugoslavia
\item \verb{[35]} romania
\item \verb{[36]} belgium
\item \verb{[37]} arabic
\item \verb{[38]} other spanish
\item \verb{[39]} non-span west indies
\item \verb{[40]} other asian
\item \verb{[41]} other european
\item \verb{[97]} american only
\item \verb{[101]} turkey
\item \verb{[202]} algeria
\item \verb{[203]} congo
\item \verb{[204]} egypt
\item \verb{[205]} ethiopia
\item \verb{[206]} kenya
\item \verb{[207]} nigeria
\item \verb{[208]} south africa
\item \verb{[209]} sudan
\item \verb{[210]} tanzania
\item \verb{[211]} uganda
\item \verb{[299]} other africa
\item \verb{[301]} south korea
\item \verb{[302]} bangladesh
\item \verb{[303]} indonesia
\item \verb{[304]} pakistan
\item \verb{[306]} thailand
\item \verb{[307]} vietnam
\item \verb{[401]} iran
\item \verb{[402]} iraq
\item \verb{[403]} israel
\item \verb{[404]} jordan
\item \verb{[405]} saudi arabia
\item \verb{[406]} syria
\item \verb{[407]} uae
\item \verb{[408]} yemen
\item \verb{[499]} other middle east
\item \verb{[501]} argentina
\item \verb{[502]} bolivia
\item \verb{[503]} brazil
\item \verb{[504]} chile
\item \verb{[505]} colombia
\item \verb{[506]} ecuador
\item \verb{[507]} french guiana
\item \verb{[508]} guyana
\item \verb{[509]} paraguay
\item \verb{[510]} peru
\item \verb{[511]} suriname
\item \verb{[512]} uruguay
\item \verb{[513]} venezuela
\item \verb{[599]} other south america
\item \verb{[601]} belize
\item \verb{[602]} costa rica
\item \verb{[603]} el salvador
\item \verb{[604]} guatemala
\item \verb{[605]} honduras
\item \verb{[606]} nicaragua
\item \verb{[607]} panama
\item \verb{[699]} other central america
\item \verb{[799]} other north america
\item \verb{[801]} cuba
\item \verb{[802]} haiti
\item \verb{[803]} dominican republic
\item \verb{[804]} jamaica
\item \verb{[899]} other caribbean
\item \verb{[901]} australia
\item \verb{[902]} papua new guinea
\item \verb{[903]} new zealand
\item \verb{[904]} samoa
\item \verb{[999]} other oceania
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnum}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}}
}
\concept{Ethnicity}
\keyword{variable}
