% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{ethnofit}
\alias{ethnofit}
\title{Eth minorities never fit in am mainstream}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethnofit
}
\details{
Question Here are some opinions some people have expressed in connection with ethnic issues in the United States. To what extent do you agree or disagree with each one?
B. Ethnic minority groups will never really fit in with mainstream American culture.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2727/vshow}{GSS Data Explorer page for \code{ethnofit}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 88 \tab 333 \tab 390 \tab 1292 \tab 614 \tab 37 \tab 11 \tab 2765 \cr
   Total \tab 0 \tab 88 \tab 333 \tab 390 \tab 1292 \tab 614 \tab 37 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelblks}},
\code{\link{feelhsps}}
}
\concept{Prejudice}
\keyword{variable}
