% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{ethnofit}
\alias{ethnofit}
\title{Eth minorities never fit in am mainstream}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethnofit
}
\details{
Question Here are some opinions some people have expressed in connection with ethnic issues in the United States. To what extent do you agree or disagree with each one?
B. Ethnic minority groups will never really fit in with mainstream American culture.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2727/vshow}{GSS Data Explorer page for \code{ethnofit}}.
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethtrads}},
\code{\link{feelasns}},
\code{\link{feelblks}},
\code{\link{feelhsps}}

Other Ethnicity: 
\code{\link{ethadapt}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethtrads}}
}
\concept{Ethnicity}
\concept{Prejudice}
\keyword{variable}
