% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{ethspkno}
\alias{ethspkno}
\title{Ban opinions harmful to other groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethspkno
}
\details{
Question B. People should not be allowed to express opinions that are harmful or offensive to members of other religious or racial groups.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1604/vshow}{GSS Data Explorer page for \code{ethspkno}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2275 \tab 71 \tab 222 \tab 86 \tab 246 \tab 71 \tab 8 \tab 13 \tab 2992 \cr
   Total \tab 2275 \tab 71 \tab 222 \tab 86 \tab 246 \tab 71 \tab 8 \tab 13 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
