% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{ethspkno}
\alias{ethspkno}
\title{Ban opinions harmful to other groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethspkno
}
\details{
Question B. People should not be allowed to express opinions that are harmful or offensive to members of other religious or racial groups.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1604/vshow}{GSS Data Explorer page for \code{ethspkno}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkok}},
\code{\link{letin}},
\code{\link{symptblk}}

Other Civil Liberties: 
\code{\link{ethspkok}}

Other Race: 
\code{\link{admirblk}},
\code{\link{ethorgs}},
\code{\link{ethspkok}},
\code{\link{symptblk}}

Other Religion: 
\code{\link{ethspkok}}

Other Split Ballots: 
\code{\link{ethspkok}}
}
\concept{Civil Liberties}
\concept{Multiculturalism}
\concept{Race}
\concept{Religion}
\concept{Split Ballots}
\keyword{variable}
