% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_612.R
\name{ethworld1}
\alias{ethworld1}
\title{From what part of the world did your ancestors come?  Europe}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethworld1
}
\details{
Question From what part of the world did your ancestors come?  Europe
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} yes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7631/vshow}{GSS Data Explorer page for \code{ethworld1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab - \tab 796 \tab 2677 \tab 544 \tab 1 \tab 14 \tab 4032 \cr
   2022 \tab - \tab 1072 \tab 2254 \tab 161 \tab 56 \tab 1 \tab 3544 \cr
   Total \tab 0 \tab 1868 \tab 4931 \tab 705 \tab 57 \tab 15 \tab 7576 \cr
}
}

\keyword{variable}
