% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{evangclx}
\alias{evangclx}
\title{Is r born again or evangelical}
\source{
General Social Survey https://gss.norc.org
}
\description{
evangclx
}
\details{
Question Would you describe yourself as a "born again" or evangelical Christian, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7523/vshow}{GSS Data Explorer page for \code{evangclx}}.
}

\keyword{variable}
