% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{evangclx}
\alias{evangclx}
\title{Is r born again or evangelical}
\source{
General Social Survey https://gss.norc.org
}
\description{
evangclx
}
\details{
Question Would you describe yourself as a "born again" or evangelical Christian, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7523/vshow}{GSS Data Explorer page for \code{evangclx}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 3199 \tab 365 \tab 358 \tab 1 \tab 103 \tab 5 \tab 1 \tab 4032 \cr
   Total \tab 3199 \tab 365 \tab 358 \tab 1 \tab 103 \tab 5 \tab 1 \tab 4032 \cr
}
}

\keyword{variable}
