% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_218.R
\name{evbrkdwn}
\alias{evbrkdwn}
\title{Ever felt going to have a nervous breakdown?}
\source{
General Social Survey https://gss.norc.org
}
\description{
evbrkdwn
}
\details{
Question Have you ever felt that you were going to have a nervous breakdown?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   B/C/- \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2066/vshow}{GSS Data Explorer page for \code{evbrkdwn}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 376 \tab 1047 \tab 10 \tab 11 \tab 2904 \cr
   2016 \tab 1745 \tab 385 \tab 724 \tab 1 \tab 12 \tab 2867 \cr
   Total \tab 3205 \tab 761 \tab 1771 \tab 11 \tab 23 \tab 5771 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo1}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{knwpatnt}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}}
}
\concept{Mental Health}
\keyword{variable}
