% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{evdiv}
\alias{evdiv}
\title{Has r ever been divorced}
\source{
General Social Survey https://gss.norc.org
}
\description{
evdiv
}
\details{
Question Were you ever divorced?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} never married
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4393/vshow}{GSS Data Explorer page for \code{evdiv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab never married \tab no answer \tab Total \cr
   1988 \tab 67 \tab 361 \tab 749 \tab 280 \tab 24 \tab 1481 \cr
   1994 \tab 1545 \tab 410 \tab 844 \tab 172 \tab 21 \tab 2992 \cr
   Total \tab 1612 \tab 771 \tab 1593 \tab 452 \tab 45 \tab 4473 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabit}},
\code{\link{divhubby}},
\code{\link{earnsmor}},
\code{\link{mawork14}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
