% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{evdwell}
\alias{evdwell}
\title{In life, r has ever been prevented from moving bc landlord believed r was gay/le}
\source{
General Social Survey https://gss.norc.org
}
\description{
evdwell
}
\details{
Question Have you ever in your life been prevented from buying or moving into a house or apartment or been Forced to move because a landlord or realtor believed you were gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3926/vshow}{GSS Data Explorer page for \code{evdwell}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1879 \tab 2 \tab 113 \tab 3 \tab 26 \tab 2023 \cr
   2012 \tab 1807 \tab 3 \tab 118 \tab - \tab 46 \tab 1974 \cr
   Total \tab 3686 \tab 5 \tab 231 \tab 3 \tab 72 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
