% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{evidu}
\alias{evidu}
\title{R ever inject drugs}
\source{
General Social Survey https://gss.norc.org
}
\description{
evidu
}
\details{
Question Have you ever, even once, taken any drugs by injection with a needle (like heroin, cocaine, amphetamines, or steroids). Do not include anything you took under a doctor’s orders.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5070/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab skipped on web \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 417 \tab 1 \tab 2265 \tab 72 \tab 62 \tab - \tab - \tab 2817 \cr
   2002 \tab 476 \tab - \tab 2203 \tab 29 \tab 57 \tab - \tab - \tab 2765 \cr
   2004 \tab 570 \tab 3 \tab 2168 \tab 15 \tab 56 \tab - \tab - \tab 2812 \cr
   2006 \tab 2096 \tab - \tab 2292 \tab 22 \tab 100 \tab - \tab - \tab 4510 \cr
   2008 \tab 240 \tab 1 \tab 1706 \tab 14 \tab 62 \tab - \tab - \tab 2023 \cr
   2010 \tab 202 \tab 2 \tab 1752 \tab 22 \tab 66 \tab - \tab - \tab 2044 \cr
   2012 \tab 238 \tab 2 \tab 1661 \tab 31 \tab 42 \tab - \tab - \tab 1974 \cr
   2014 \tab 188 \tab 1 \tab 2247 \tab 38 \tab 64 \tab - \tab - \tab 2538 \cr
   2016 \tab 1096 \tab - \tab 1694 \tab 19 \tab 58 \tab - \tab - \tab 2867 \cr
   2018 \tab 942 \tab - \tab 1350 \tab 15 \tab 41 \tab - \tab - \tab 2348 \cr
   2021 \tab 1708 \tab 36 \tab 2198 \tab 2 \tab 61 \tab 27 \tab - \tab 4032 \cr
   2022 \tab 1929 \tab 23 \tab 1534 \tab 1 \tab 44 \tab 13 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 48218 \tab 69 \tab 23070 \tab 280 \tab 713 \tab 40 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest2}},
\code{\link{idu30}},
\code{\link{relatsex}}

Other Drug Use And Addiction: 
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{idu30}}
}
\concept{Drug Use And Addiction}
\concept{High Risk Behaviors}
\keyword{variable}
