% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{evmhp}
\alias{evmhp}
\title{Ever had mental health problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
evmhp
}
\details{
Question (IF NO TO EVBRKDWN, ASK EVMHP) Have you ever felt you had a mental health problem?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2078/vshow}{GSS Data Explorer page for \code{evmhp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1837 \tab 76 \tab 977 \tab 1 \tab 13 \tab 2904 \cr
   Total \tab 1837 \tab 76 \tab 977 \tab 1 \tab 13 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdtime}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
