% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{evngelze}
\alias{evngelze}
\title{Important to encourage non-christians to accept jesus as savior}
\source{
General Social Survey https://gss.norc.org
}
\description{
evngelze
}
\details{
Question It is very important for me personally to encourage non-Christians to accept Jesus Christ as their Savior.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7531/vshow}{GSS Data Explorer page for \code{evngelze}}.
}

\keyword{variable}
