% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{evolved1}
\alias{evolved1}
\title{Humans developed from earlier species: true or false}
\source{
General Social Survey https://gss.norc.org
}
\description{
evolved1
}
\details{
Question According to the theory of evolution, human beings, as we know them today, developed from earlier species of animals. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3549/vshow}{GSS Data Explorer page for \code{evolved1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab can't choose \tab no answer \tab Total \cr
   2012 \tab 1489 \tab 340 \tab 94 \tab 46 \tab 5 \tab 1974 \cr
   Total \tab 1489 \tab 340 \tab 94 \tab 46 \tab 5 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
