% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{evolved2}
\alias{evolved2}
\title{Elephants evolved from earlier species: true or false}
\source{
General Social Survey https://gss.norc.org
}
\description{
evolved2
}
\details{
Question For each statement that I read, please tell me if it is true or false. If you don't know or aren't sure, just tell me so, and we will skip to the next question. Remember true, false, or don't know.
I3. Elephants, as we know them today, descended from earlier species of animals. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6303/vshow}{GSS Data Explorer page for \code{evolved2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2181 \tab 498 \tab 84 \tab 100 \tab 4 \tab 2867 \cr
   2018 \tab 1763 \tab 408 \tab 84 \tab 90 \tab 3 \tab 2348 \cr
   Total \tab 3944 \tab 906 \tab 168 \tab 190 \tab 7 \tab 5215 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
