% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{evunemp}
\alias{evunemp}
\title{Ever been unemployed two weeks +}
\source{
General Social Survey https://gss.norc.org
}
\description{
evunemp
}
\details{
Question Since your first full-time job have you ever been unemployed and looking for work for as long as two weeks?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{7} never full-time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1272/vshow}{GSS Data Explorer page for \code{evunemp}}.
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{numunemp}},
\code{\link{orgfin}},
\code{\link{totunemp}},
\code{\link{wrkyears}}
}
\concept{Work Organizations}
\keyword{variable}
