% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_152.R
\name{excelart}
\alias{excelart}
\title{Artistic excellence found in folk art}
\source{
General Social Survey https://gss.norc.org
}
\description{
excelart
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
H. Artistic excellence can be found in popular and folk culture just as much as in the fine arts.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1429/vshow}{GSS Data Explorer page for \code{excelart}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 182 \tab 1208 \tab 67 \tab 6 \tab 130 \tab 13 \tab 1606 \cr
   Total \tab 0 \tab 182 \tab 1208 \tab 67 \tab 6 \tab 130 \tab 13 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{hosthome}},
\code{\link{impfinan}}
}
\concept{Culture}
\keyword{variable}
