% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{excldimm}
\alias{excldimm}
\title{America should exclude illegal immigrants}
\source{
General Social Survey https://gss.norc.org
}
\description{
excldimm
}
\details{
Question How much do you agree or disagree with the following statement?
America should take stronger measures to exclude illegal immigrants.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4797/vshow}{GSS Data Explorer page for \code{excldimm}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 592 \tab 395 \tab 179 \tab 86 \tab 27 \tab 68 \tab 20 \tab 2904 \cr
   2004 \tab 1596 \tab 394 \tab 434 \tab 214 \tab 113 \tab 32 \tab 28 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 287 \tab 449 \tab 205 \tab 227 \tab 49 \tab 53 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 1273 \tab 1278 \tab 598 \tab 426 \tab 108 \tab 149 \tab 25 \tab 8254 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{econgrn}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{ldcgrn}},
\code{\link{nukeacc}}
}
\concept{ISSP Environment}
\keyword{variable}
