% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{exjobsat}
\alias{exjobsat}
\title{Satisfaction with past job}
\source{
General Social Survey https://gss.norc.org
}
\description{
exjobsat
}
\details{
Question (HAND CARD D20)
How satisfied were you in your last job?
}
\section{Values}{

\itemize{
\item \code{1} completely satisfied
\item \code{2} very satisfied
\item \code{3} fairly satisfied
\item \code{4} neither satisfied nor dissatisfied
\item \code{5} fairly dissatisfied
\item \code{6} very dissatisfied
\item \code{7} completely dissatisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6341/vshow}{GSS Data Explorer page for \code{exjobsat}}.
}

\seealso{
Other Work Orientation: 
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
