% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{expdesgn}
\alias{expdesgn}
\title{Better way to test drug btw control and non-control}
\source{
General Social Survey https://gss.norc.org
}
\description{
expdesgn
}
\details{
Question Now, please think about this situation. Two scientists want to know if a certain drug is effective against high blood pressure. The first scientist wants to give the drug to one thousand people with high blood pressure and see how many of them experience lower blood pressure levels. The second scientist wants to give the drug to five hundred people with high blood pressure, and not give the drug to another five hundred people with high blood pressure, and see how many in both groups experience lower blood pressure levels. Which is the better way to test this drug?
}
\section{Values}{

\itemize{
\item \code{1} all 1000 get the drug
\item \code{2} 500 get the drug; 500 don't
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3330/vshow}{GSS Data Explorer page for \code{expdesgn}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{exptext}},
\code{\link{odds1}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
