% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{exptext}
\alias{exptext}
\title{Why is it better to test drug this way}
\source{
General Social Survey https://gss.norc.org
}
\description{
exptext
}
\details{
Question Why is it better to test the drug this way?
}
\section{Values}{

\itemize{
\item \verb{[1]} correct - control group
\item \verb{[2]} correct - vague reason
\item \verb{[3]} correct - wrong reason
\item \verb{[4]} correct - don't know
\item \verb{[5]} wrong - reservations about control group
\item \verb{[6]} wrong - reservations about sample
\item \verb{[7]} wrong - other reason
\item \verb{[8]} wrong - don't know
\item \verb{[9]} no answer to closed ended question
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3331/vshow}{GSS Data Explorer page for \code{exptext}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab correct - control group \tab correct - vague reason \tab correct - wrong reason \tab correct - don't know \tab wrong - reservations about control group \tab wrong - reservations about sample \tab wrong - other reason \tab wrong - don't know \tab no answer to closed ended question \tab uncodeable \tab Total \cr
   2006 \tab 2646 \tab 300 \tab 450 \tab 682 \tab 35 \tab 27 \tab 157 \tab 106 \tab 21 \tab 84 \tab 2 \tab 4510 \cr
   2008 \tab 629 \tab 199 \tab 362 \tab 513 \tab 14 \tab 37 \tab 120 \tab 106 \tab 8 \tab 2 \tab 33 \tab 2023 \cr
   2010 \tab 1364 \tab 240 \tab 114 \tab 193 \tab 17 \tab 10 \tab 62 \tab 21 \tab 6 \tab 17 \tab - \tab 2044 \cr
   2012 \tab 1038 \tab 154 \tab 144 \tab 460 \tab 33 \tab 23 \tab 52 \tab 55 \tab 4 \tab 11 \tab - \tab 1974 \cr
   2014 \tab 1366 \tab 194 \tab 394 \tab 347 \tab 12 \tab 20 \tab 93 \tab 86 \tab 2 \tab 24 \tab - \tab 2538 \cr
   2016 \tab 1526 \tab 264 \tab 428 \tab 364 \tab 34 \tab 36 \tab 104 \tab 80 \tab 12 \tab 19 \tab - \tab 2867 \cr
   2018 \tab 1212 \tab 288 \tab 281 \tab 303 \tab 28 \tab 44 \tab 104 \tab 57 \tab 17 \tab 14 \tab - \tab 2348 \cr
   Total \tab 9781 \tab 1639 \tab 2173 \tab 2862 \tab 173 \tab 197 \tab 692 \tab 511 \tab 70 \tab 171 \tab 35 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{odds1}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
