% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{extr2021}
\alias{extr2021}
\title{Extra pay in 2021}
\source{
General Social Survey https://gss.norc.org
}
\description{
extr2021
}
\details{
Question Did you receive any of these performance-based payments in the year 2021? Performance-based pay means individual or group bonuses, commissions, or any type of profit-sharing.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7850/vshow}{GSS Data Explorer page for \code{extr2021}}.
}

\keyword{variable}
