% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{extra1}
\alias{extra1}
\title{Other problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
extra1
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
I. Aside from the things we've already covered, what other important problems did you or members of your household have during the last 12 months?
PROBE FOR TWO MENTIONS. RECORD FIRST TWO MENTIONS BELOW.
\enumerate{
\item First mention.
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5194/vshow}{GSS Data Explorer page for \code{extra1}}.
}

\seealso{
Other Negative Life Events: 
\code{\link{extra2}},
\code{\link{oth5}}

Other Life Events: 
\code{\link{extra2}}
}
\concept{Life Events}
\concept{Negative Life Events}
\keyword{variable}
