% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{extraern}
\alias{extraern}
\title{How much earned from additional jobs in past yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
extraern
}
\details{
Question (HAND CARD D19)
Over the entire 12 months, how much did you earn from your additional job(s) in total compared with your main job?
From my additional job(s), I earned...
}
\section{Values}{

\itemize{
\item \code{1} much less than from main job
\item \code{2} less than from main job
\item \code{3} about the same as from main job
\item \code{4} more than from main job
\item \code{5} much more than from main job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6338/vshow}{GSS Data Explorer page for \code{extraern}}.
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
