% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{extraern}
\alias{extraern}
\title{How much earned from additional jobs in past yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
extraern
}
\details{
Question (HAND CARD D19)
Over the entire 12 months, how much did you earn from your additional job(s) in total compared with your main job?
From my additional job(s), I earned...
}
\section{Values}{

\itemize{
\item \verb{[1]} much less than from main job
\item \verb{[2]} less than from main job
\item \verb{[3]} about the same as from main job
\item \verb{[4]} more than from main job
\item \verb{[5]} much more than from main job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6338/vshow}{GSS Data Explorer page for \code{extraern}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much less than from main job \tab less than from main job \tab about the same as from main job \tab more than from main job \tab much more than from main job \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2567 \tab 152 \tab 88 \tab 24 \tab 13 \tab 11 \tab 4 \tab 8 \tab 2867 \cr
   Total \tab 2567 \tab 152 \tab 88 \tab 24 \tab 13 \tab 11 \tab 4 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
