% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_321.R
\name{extrapay}
\alias{extrapay}
\title{Eligible for performance based pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
extrapay
}
\details{
Question In your job are you eligible for any type of performance-based pay, such as individual or group bonuses or any type of profit-sharing?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2855/vshow}{GSS Data Explorer page for \code{extrapay}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 597 \tab 1164 \tab 19 \tab 16 \tab - \tab 2765 \cr
   2006 \tab 2778 \tab 644 \tab 1046 \tab 25 \tab 17 \tab - \tab 4510 \cr
   2014 \tab 1765 \tab 261 \tab 502 \tab 7 \tab 3 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 500 \tab 898 \tab 18 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 2569 \tab 322 \tab 599 \tab 37 \tab 7 \tab 10 \tab 3544 \cr
   Total \tab 8981 \tab 2324 \tab 4209 \tab 106 \tab 75 \tab 10 \tab 15705 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{deptperf}},
\code{\link{extr2001}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{seecowrk}},
\code{\link{yearval}}
}
\concept{Employee Compensation}
\keyword{variable}
