% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{extraval1}
\alias{extraval1}
\title{Estimate of bonus payments last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
extraval1
}
\details{
Question What was the approximate total dollar value of these various payments in 2021?
}
\section{Values}{

\itemize{
\item \verb{[1]} less than $100
\item \verb{[2]} $100 to $299
\item \verb{[3]} $300 to $499
\item \verb{[4]} $500 to $999
\item \verb{[5]} $1,000 to $1,499
\item \verb{[6]} $1,500 to $2,499
\item \verb{[7]} $2,500 to $3,999
\item \verb{[8]} $4,000 to $6,999
\item \verb{[9]} $7,000 to $14,999
\item \verb{[10]} $15,000 to $49,999
\item \verb{[11]} $50,000 to $99,999
\item \verb{[12]} over $100,000
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7851/vshow}{GSS Data Explorer page for \code{extraval1}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab 10 \tab 11 \tab 12 \tab Total \cr
   2022 \tab 3354 \tab 7 \tab 15 \tab 15 \tab 16 \tab 22 \tab 14 \tab 14 \tab 18 \tab 30 \tab 24 \tab 7 \tab 8 \tab 3544 \cr
   Total \tab 3354 \tab 7 \tab 15 \tab 15 \tab 16 \tab 22 \tab 14 \tab 14 \tab 18 \tab 30 \tab 24 \tab 7 \tab 8 \tab 3544 \cr
}
}

\keyword{variable}
