% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{extrawrk}
\alias{extrawrk}
\title{Did you work multiple jobs in past yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
extrawrk
}
\details{
Question (HAND CARD D18)
Over the past 12 months, in addition to your main job, have you done any other work for pay?
}
\section{Values}{

\itemize{
\item \code{1} yes, during all of that period
\item \code{2} yes, during most of that period
\item \code{3} yes, during some of that period
\item \code{4} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6337/vshow}{GSS Data Explorer page for \code{extrawrk}}.
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
