% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{extrawrk}
\alias{extrawrk}
\title{Did you work multiple jobs in past yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
extrawrk
}
\details{
Question (HAND CARD D18)
Over the past 12 months, in addition to your main job, have you done any other work for pay?
}
\section{Values}{

\itemize{
\item \code{1} yes, during all of that period
\item \code{2} yes, during most of that period
\item \code{3} yes, during some of that period
\item \code{4} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6337/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab yes, during all of that period \tab yes, during most of that period \tab yes, during some of that period \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1927 \tab 640 \tab 6 \tab 83 \tab 45 \tab 166 \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 67906 \tab 640 \tab 6 \tab 83 \tab 45 \tab 166 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
