% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{extrawrk}
\alias{extrawrk}
\title{Did you work multiple jobs in past yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
extrawrk
}
\details{
Question (HAND CARD D18)
Over the past 12 months, in addition to your main job, have you done any other work for pay?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, during all of that period
\item \verb{[2]} yes, during most of that period
\item \verb{[3]} yes, during some of that period
\item \verb{[4]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6337/vshow}{GSS Data Explorer page for \code{extrawrk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, during all of that period \tab yes, during most of that period \tab yes, during some of that period \tab no \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 83 \tab 45 \tab 166 \tab 640 \tab 6 \tab 2867 \cr
   Total \tab 1927 \tab 83 \tab 45 \tab 166 \tab 640 \tab 6 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
