% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{fair5}
\alias{fair5}
\title{People fair or not}
\source{
General Social Survey https://gss.norc.org
}
\description{
fair5
}
\details{
Question B2. Generally speaking, do you think that most people would try to take advantage of you if they got the chance, or would they try to be fair?
Please tell me what you think, where 1 means most people would try to take advantage of you and 5 means that most people would try to be fair.
}
\section{Values}{

\itemize{
\item \code{1} most people would try to take advantage
\item \code{5} most people would try to be fair
\item \code{6} 6 can't choose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4801/vshow}{GSS Data Explorer page for \code{fair5}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}

Other People: 
\code{\link{trust5}}
}
\concept{ISSP Environment}
\concept{People}
\keyword{variable}
