% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{fairearn}
\alias{fairearn}
\title{How fair is what r earn on the job}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairearn
}
\details{
Question How fair is what you earn on your job in comparison to others doing the same type of work you do?
}
\section{Values}{

\itemize{
\item \code{1} much less than you deserve
\item \code{2} somewhat less than you deserve
\item \code{3} about as much as you deserve
\item \code{4} somewhat more than you deserve
\item \code{5} much more than you deserve
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2816/vshow}{GSS Data Explorer page for \code{fairearn}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Income}
\concept{Quality of Working Life}
\keyword{variable}
