% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{fairhwrk}
\alias{fairhwrk}
\title{Division of work unfair to you or spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairhwrk
}
\details{
Question We're interested in knowing how much of the work around your home is done by different people. We don't mean taking care of children - just regular work around the house, like cooking, grocery shopping and doing little repair jobs.
C. How fair do you feel the division of work around the house is in your household? Would you say that it is fair to both you and your spouse or partner, unfair to you, or unfair to your spouse or partner?
}
\section{Values}{

\itemize{
\item \code{1} fair to both
\item \code{2} unfair to me
\item \code{3} unfair to my spouse or partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1662/vshow}{GSS Data Explorer page for \code{fairhwrk}}.
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{balwkfam}},
\code{\link{children}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{earnmore}},
\code{\link{emoteoth}},
\code{\link{famlife}},
\code{\link{febear}},
\code{\link{fehire}},
\code{\link{feimp}},
\code{\link{fejobaff}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{flextime}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{haprom}},
\code{\link{hmemaker}},
\code{\link{livewith}},
\code{\link{manprof}},
\code{\link{marfree}},
\code{\link{mebear}},
\code{\link{men}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{parleave}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{rhmewrk}},
\code{\link{romance}},
\code{\link{rushed}},
\code{\link{selffrst}},
\code{\link{sexhar}},
\code{\link{sexpromo}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{twoincs}},
\code{\link{willwed1}},
\code{\link{willwed2}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{wrkclass}},
\code{\link{yourself}}
}
\concept{Gender}
\keyword{variable}
