% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{fairnv}
\alias{fairnv}
\title{People fair or try to take advantage (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairnv
}
\details{
Question Do you think most people would try to take advantage of you if they got a chance, or would they try to be fair?
}
\section{Values}{

\itemize{
\item \code{1} would take advantage of you
\item \code{2} would try to be fair
\item \code{3} depends (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7612/vshow}{GSS Data Explorer page for \code{fairnv}}.
}

\keyword{variable}
