% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_251.R
\name{fairwhts}
\alias{fairwhts}
\title{Whites committed to fair and equal treatment of}
\source{
General Social Survey https://gss.norc.org
}
\description{
fairwhts
}
\details{
Question In the next statement a score of 1 means that you think almost all of the people in the group have a "commitment to the fair and equal treatment of all groups in society."  A score of 7 means that you think almost everyone in the group "lacks a commitment to the fair and equal treatment of all groups in society."  A score of 4 means that you think that the group is not towards one end or the other, and of course you may choose any number in between that comes closest to where you think people in the group stand.
A. Where would you rate Whites in general on this scale?
}
\section{Values}{

\itemize{
\item \code{1} commitment to fair and equal treatment
\item \code{7} lacks commitment to fair and equal treatment
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2266/vshow}{GSS Data Explorer page for \code{fairwhts}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
