% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{fambiz}
\alias{fambiz}
\title{What kind of business is this?}
\source{
General Social Survey https://gss.norc.org
}
\description{
fambiz
}
\details{
Question What kind of business is this?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} retail sales
\item \verb{[3]} wholesale sales
\item \verb{[4]} services
\item \verb{[5]} transportation
\item \verb{[6]} construction
\item \verb{[7]} finance, insrnce or real est
\item \verb{[8]} farming
\item \verb{[9]} other specify
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1878/vshow}{GSS Data Explorer page for \code{fambiz}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab manufacturing \tab retail sales \tab wholesale sales \tab services \tab transportation \tab construction \tab finance, insrnce or real est \tab farming \tab other specify \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2841 \tab 4 \tab 6 \tab 1 \tab 11 \tab 5 \tab 9 \tab 1 \tab 6 \tab 8 \tab 1 \tab 11 \tab 2904 \cr
   Total \tab 2841 \tab 4 \tab 6 \tab 1 \tab 11 \tab 5 \tab 9 \tab 1 \tab 6 \tab 8 \tab 1 \tab 11 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
