% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{fambudgt}
\alias{fambudgt}
\title{How couples monitor budget}
\source{
General Social Survey https://gss.norc.org
}
\description{
fambudgt
}
\details{
Question How do you and your spouse/partner organize the income that one or both of you receive?  Please choose the option that comes closest.
}
\section{Values}{

\itemize{
\item \code{1} i manage, give prtnr share
\item \code{2} prntnr manages, gives me share
\item \code{3} pool all money, take out what we need
\item \code{4} pool some money, keep some separate
\item \code{5} keep money separate
\item \code{6} not married
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4417/vshow}{GSS Data Explorer page for \code{fambudgt}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab i manage, give prtnr share \tab prntnr manages, gives me share \tab pool all money, take out what we need \tab pool some money, keep some separate \tab keep money separate \tab not married \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 95 \tab 62 \tab 483 \tab 80 \tab 76 \tab 612 \tab 1 \tab 38 \tab 2992 \cr
   2002 \tab 1594 \tab 82 \tab 83 \tab 366 \tab 69 \tab 86 \tab 456 \tab 8 \tab 21 \tab 2765 \cr
   2012 \tab 1274 \tab 111 \tab 90 \tab 338 \tab 67 \tab 81 \tab - \tab 4 \tab 9 \tab 1974 \cr
   Total \tab 4413 \tab 288 \tab 235 \tab 1187 \tab 216 \tab 243 \tab 1068 \tab 13 \tab 68 \tab 7731 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
