% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{famfinan}
\alias{famfinan}
\title{How you and your spouse organize your finances}
\source{
General Social Survey https://gss.norc.org
}
\description{
famfinan
}
\details{
Question Now I''d like to ask you about how you and your (husband/wife) organize your finances. Which of the following comes closest to describing the system that you and your (husband/wife) use:
}
\section{Values}{

\itemize{
\item \code{1} wife manages all the money
\item \code{2} husband manages all the money
\item \code{3} wife has a housekeeping allowance
\item \code{4} all the money is pooled
\item \code{5} some of the money is pooled
\item \code{6} keep finances completely separate
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1880/vshow}{GSS Data Explorer page for \code{famfinan}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab wife manages all the money \tab husband manages all the money \tab wife has a housekeeping allowance \tab all the money is pooled \tab some of the money is pooled \tab keep finances completely separate \tab no answer \tab Total \cr
   1996 \tab 2196 \tab 200 \tab 61 \tab 59 \tab 260 \tab 91 \tab 32 \tab 5 \tab 2904 \cr
   Total \tab 2196 \tab 200 \tab 61 \tab 59 \tab 260 \tab 91 \tab 32 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
