% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{famgen_exp}
\alias{famgen_exp}
\title{Number of family generations in household (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
famgen_exp
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} one generation
\item \code{2} two generations, children
\item \code{3} two generations, parents
\item \code{4} two generations, grandchild(ren)
\item \code{5} three generations, child(ren), grandchildren
\item \code{6} three generations, child(ren), parent
\item \code{7} four generations
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
