% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_668.R
\name{famgen_exp}
\alias{famgen_exp}
\title{Number of family generations in household (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
famgen_exp
}
\details{
Question Recoded from experimental Household Composition variables ADULTS_EXP to RNRLTU18
As of 2024, FAMGEN_EXP has been recoded to only account for 3+ generations. This has been retroactively applied to all previous years of FAMGEN_EXP. The original FAMGEN containing data in 2022 has been renamed FAMGEN_EXP_22.
}
\section{Values}{

\itemize{
\item \code{1} one generation
\item \code{2} two generations, children
\item \code{3} two generations, parents
\item \code{4} two generations, grandchild(ren)
\item \code{5} three generations, child(ren), grandchildren
\item \code{6} three generations, child(ren), parent
\item \code{7} four generations
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8034/vshow}{GSS Data Explorer page for \code{famgen_exp}}.
}

\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
