% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_251.R
\name{famhsps}
\alias{famhsps}
\title{Hispanics committed to strong families}
\source{
General Social Survey https://gss.norc.org
}
\description{
famhsps
}
\details{
Question In the next statement a score of 1 means that you think almost all of the people in the group have a "commitment to strong families."  A score of 7 means that you think almost everyone in the group "lacks a commitment to strong families."  A score of 4 means that you think that the group is not towards one end or the other, and of course you may choose any number in between that comes closest to where you think people in the group stand.
D. Hispanics or Latin Americans?
}
\section{Values}{

\itemize{
\item \verb{[1]} commitment to strong families
\item \verb{[7]} lacks commitment to strong families
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2264/vshow}{GSS Data Explorer page for \code{famhsps}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab commitment to strong families \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab lacks commitment to strong families \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 167 \tab 276 \tab 276 \tab 369 \tab 118 \tab 36 \tab 17 \tab 125 \tab 14 \tab 2817 \cr
   Total \tab 1419 \tab 167 \tab 276 \tab 276 \tab 369 \tab 118 \tab 36 \tab 17 \tab 125 \tab 14 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famjews}},
\code{\link{famwhts}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
