% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{family16sex}
\alias{family16sex}
\title{Sex of two parents living with when 16 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
family16sex
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} mother and father
\item \code{2} two mothers
\item \code{3} two fathers
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
