% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{family16sex}
\alias{family16sex}
\title{Sex of two parents living with when 16 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
family16sex
}
\details{
Question Were you living with your own two parents around the time you were 16? With whom were you living around that time?
}
\section{Values}{

\itemize{
\item \code{1} mother and father
\item \code{2} two mothers
\item \code{3} two fathers
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8107/vshow}{GSS Data Explorer page for \code{family16sex}}.
}

\seealso{
Other Core: 
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{racdif5x}},
\code{\link{sexnow2}}
}
\concept{Core}
\keyword{variable}
