% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{famintjb}
\alias{famintjb}
\title{How often r felt: demand of family life interfere with job}
\source{
General Social Survey https://gss.norc.org
}
\description{
famintjb
}
\details{
Question How often do you feel that the demands of your family life interfere with your job?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} hardly ever
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4615/vshow}{GSS Data Explorer page for \code{famintjb}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 13 \tab 41 \tab 250 \tab 387 \tab 325 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 17 \tab 51 \tab 257 \tab 334 \tab 275 \tab 6 \tab 2867 \cr
   Total \tab 5419 \tab 30 \tab 92 \tab 507 \tab 721 \tab 600 \tab 8 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
