% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{famintjb}
\alias{famintjb}
\title{How often r felt: demand of family life interfere with job}
\source{
General Social Survey https://gss.norc.org
}
\description{
famintjb
}
\details{
Question How often do you feel that the demands of your family life interfere with your job?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4615/vshow}{GSS Data Explorer page for \code{famintjb}}.
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}

Other Job: 
\code{\link{dailywrk}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Job}
\concept{Work Orientation}
\keyword{variable}
