% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{famlife}
\alias{famlife}
\title{How successful in family life}
\source{
General Social Survey https://gss.norc.org
}
\description{
famlife
}
\details{
Question The next few questions are about your success in different areas of life. First, how successful do you feel in your family life? Do you feel...
}
\section{Values}{

\itemize{
\item \code{1} not at all successful
\item \code{2} not very successful
\item \code{3} somewhat successful
\item \code{4} very successful
\item \code{5} completely successful
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1664/vshow}{GSS Data Explorer page for \code{famlife}}.
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{earnmore}},
\code{\link{fairhwrk}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}

Other Family: 
\code{\link{balwkfam}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}}

Other Success: 
\code{\link{balwkfam}},
\code{\link{worklife}}
}
\concept{Family}
\concept{Gender}
\concept{Success}
\keyword{variable}
