% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_225.R
\name{fammhneg}
\alias{fammhneg}
\title{R's family's negative attitudes about mh problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
fammhneg
}
\details{
Question Thinking about your family, to what extent do they hold negative attitudes about people with mental health problems?
}
\section{Values}{

\itemize{
\item \code{1} very much
\item \code{2} somewhat
\item \code{3} not very much
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7145/vshow}{GSS Data Explorer page for \code{fammhneg}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very much \tab somewhat \tab not very much \tab not at all \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 31 \tab 192 \tab 390 \tab 477 \tab 31 \tab 52 \tab 2348 \cr
   Total \tab 1175 \tab 31 \tab 192 \tab 390 \tab 477 \tab 31 \tab 52 \tab 2348 \cr
}
}

\concept{Mental Health}
\keyword{variable}
