% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{famresp}
\alias{famresp}
\title{Importance of family resps. to pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
famresp
}
\details{
Question Think of two people doing the same kind of work. What do you personally think should be important in deciding how much to pay them?
Looking at the things below, please write ''1'' next to the thing that you think should be most important. Then write ''2'' next to the thing that you think should be next most important, and ''3'' next to the thing that you think should be third most important. Leave the other boxes blank.
G. The employee''s family responsibilities
}
\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} 2nd most impt
\item \code{3} 3rd most impt
\item \code{4} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4502/vshow}{GSS Data Explorer page for \code{famresp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most impt \tab 2nd most impt \tab 3rd most impt \tab not chosen \tab don't know \tab no answer \tab Total \cr
   1989 \tab 84 \tab 5 \tab 19 \tab 42 \tab 1019 \tab 53 \tab 315 \tab 1537 \cr
   Total \tab 84 \tab 5 \tab 19 \tab 42 \tab 1019 \tab 53 \tab 315 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{bosswrks}},
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{hiinc}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}},
\code{\link{strngun}}
}
\concept{Work Orientation}
\keyword{variable}
