% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{famrnk}
\alias{famrnk}
\title{Rank in society of r}
\source{
General Social Survey https://gss.norc.org
}
\description{
famrnk
}
\details{
Question Now we would like to know where you and other people fit in. Please look at the diagram you think best describes America today, the one you chose first:
A. Where would you say you and your family actually are - at the top, somewhere in the middle, or toward the bottom?
}
\section{Values}{

\itemize{
\item \code{1} top
\item \code{7} bottom
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4338/vshow}{GSS Data Explorer page for \code{famrnk}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}

Other Social Class: 
\code{\link{execrnk}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\concept{Social Class}
\keyword{variable}
