% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{famsuffr}
\alias{famsuffr}
\title{Family life suffers if mom works f-t}
\source{
General Social Survey https://gss.norc.org
}
\description{
famsuffr
}
\details{
Question Do you agree or disagree ...
C. All in all, family life suffers when the woman has a full-time job.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
   2022 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4345/vshow}{GSS Data Explorer page for \code{famsuffr}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fejobind}},
\code{\link{hapifwrk}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{kidsuffr}},
\code{\link{mawrkwrm}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
