% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{famwkbst}
\alias{famwkbst}
\title{Best way to org family work life with under school age child}
\source{
General Social Survey https://gss.norc.org
}
\description{
famwkbst
}
\details{
Question Consider a family with a child under school age. What, in your opinion, is the best way for them to organize their family and work life? Should…
}
\section{Values}{

\itemize{
\item \verb{[1]} the mother stays at home and the father works full-time
\item \verb{[2]} the mother works part-time and the father works full-time
\item \verb{[3]} both the mother and father work full-time
\item \verb{[4]} both the mother and father work part-time
\item \verb{[5]} the father works part-time and the mother works full-time
\item \verb{[6]} the father stays at home and the mother works full-time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4467/vshow}{GSS Data Explorer page for \code{famwkbst}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab the mother stays at home and the father works full-time \tab the mother works part-time and the father works full-time \tab both the mother and father work full-time \tab both the mother and father work part-time \tab the father works part-time and the mother works full-time \tab the father stays at home and the mother works full-time \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 379 \tab 416 \tab 113 \tab 60 \tab 2 \tab 7 \tab 316 \tab 9 \tab - \tab 1974 \cr
   2022 \tab 2383 \tab 219 \tab 252 \tab 231 \tab 94 \tab 2 \tab 4 \tab 341 \tab 7 \tab 11 \tab 3544 \cr
   Total \tab 3055 \tab 598 \tab 668 \tab 344 \tab 154 \tab 4 \tab 11 \tab 657 \tab 16 \tab 11 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
