% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{famwklst}
\alias{famwklst}
\title{Wor'st way to org family work life with under school age child}
\source{
General Social Survey https://gss.norc.org
}
\description{
famwklst
}
\details{
Question And, in your opinion, which of these options would be the least desirable?
}
\section{Values}{

\itemize{
\item \verb{[1]} the mother stays at home and the father works full-time
\item \verb{[2]} the mother works part-time and the father works full-time
\item \verb{[3]} both the mother and father work full-time
\item \verb{[4]} both the mother and father work part-time
\item \verb{[5]} the father works part-time and the mother works full-time
\item \verb{[6]} the father stays at home and the mother works full-time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4468/vshow}{GSS Data Explorer page for \code{famwklst}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab the mother stays at home and the father works full-time \tab the mother works part-time and the father works full-time \tab both the mother and father work full-time \tab both the mother and father work part-time \tab the father works part-time and the mother works full-time \tab the father stays at home and the mother works full-time \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 59 \tab 24 \tab 378 \tab 139 \tab 51 \tab 391 \tab 249 \tab 11 \tab - \tab 1974 \cr
   2022 \tab 2384 \tab 68 \tab 20 \tab 304 \tab 95 \tab 34 \tab 250 \tab 369 \tab 9 \tab 11 \tab 3544 \cr
   Total \tab 3056 \tab 127 \tab 44 \tab 682 \tab 234 \tab 85 \tab 641 \tab 618 \tab 20 \tab 11 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
