% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_389.R
\name{farming}
\alias{farming}
\title{How scientific is farming}
\source{
General Social Survey https://gss.norc.org
}
\description{
farming
}
\details{
Question A. Farming. Is farming very scientific, pretty scientific, not too scientific, or not scientific at all?
}
\section{Values}{

\itemize{
\item \code{1} very scientific
\item \code{2} pretty scientific
\item \code{3} not too scientific
\item \code{4} not scientific at all
\item \code{5} haven't heard of it (volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3537/vshow}{GSS Data Explorer page for \code{farming}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{engint}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{journlsm}}

Other Employment: 
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{engint}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{journlsm}}

Other Job: 
\code{\link{engbrng}},
\code{\link{engbtr}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{engint}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{journlsm}}

Other Occupation: 
\code{\link{journlsm}}

Other Work: 
\code{\link{journlsm}}
}
\concept{Employment}
\concept{Job}
\concept{Occupation}
\concept{Science Knowledge & Attitudes}
\concept{Work}
\keyword{variable}
