% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{fatigue}
\alias{fatigue}
\title{In the past 7 days, r's average fatigue}
\source{
General Social Survey https://gss.norc.org
}
\description{
fatigue
}
\details{
Question In the past seven days, how would you rate your fatigue on average?
}
\section{Values}{

\itemize{
\item \code{1} none
\item \code{2} mild
\item \code{3} moderate
\item \code{4} severe
\item \code{5} very severe
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7170/vshow}{GSS Data Explorer page for \code{fatigue}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab none \tab mild \tab moderate \tab severe \tab very severe \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 409 \tab 898 \tab 795 \tab 184 \tab 37 \tab 7 \tab 18 \tab - \tab 2348 \cr
   2021 \tab 389 \tab 684 \tab 1491 \tab 1124 \tab 253 \tab 63 \tab 2 \tab 1 \tab 25 \tab 4032 \cr
   Total \tab 389 \tab 1093 \tab 2389 \tab 1919 \tab 437 \tab 100 \tab 9 \tab 19 \tab 25 \tab 6380 \cr
}
}

\keyword{variable}
