% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{fatigue}
\alias{fatigue}
\title{In the past 7 days, r's average fatigue}
\source{
General Social Survey https://gss.norc.org
}
\description{
fatigue
}
\details{
Question In the past seven days, how would you rate your fatigue on average?
}
\section{Values}{

\itemize{
\item \code{1} none
\item \code{2} mild
\item \code{3} moderate
\item \code{4} severe
\item \code{5} very severe
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7170/vshow}{GSS Data Explorer page for \code{fatigue}}.
}

\keyword{variable}
