% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_286.R
\name{fbaccess}
\alias{fbaccess}
\title{Permission to access r's facebook account}
\source{
General Social Survey https://gss.norc.org
}
\description{
fbaccess
}
\details{
Question Permission to access respondent's facebook account
}
\section{Values}{

\itemize{
\item \verb{[1]} permission granted
\item \verb{[2]} permission denied
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7427/vshow}{GSS Data Explorer page for \code{fbaccess}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab permission granted \tab permission denied \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1903 \tab 317 \tab 623 \tab 2 \tab 22 \tab 2867 \cr
   Total \tab 1903 \tab 317 \tab 623 \tab 2 \tab 22 \tab 2867 \cr
}
}

\keyword{variable}
