% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{fear}
\alias{fear}
\title{Afraid to walk at night in neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
fear
}
\details{
Question Is there any area right around here--that is, within a mile--where you would be afraid to walk alone at night?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
   A/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2022, 2024 \cr
   Full \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/675/vshow}{GSS Data Explorer page for \code{fear}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1973 \tab - \tab 611 \tab 877 \tab 8 \tab 8 \tab - \tab 1504 \cr
   1974 \tab - \tab 663 \tab 809 \tab 8 \tab 4 \tab - \tab 1484 \cr
   1976 \tab - \tab 657 \tab 835 \tab 5 \tab 2 \tab - \tab 1499 \cr
   1977 \tab - \tab 686 \tab 834 \tab 9 \tab 1 \tab - \tab 1530 \cr
   1980 \tab - \tab 629 \tab 827 \tab 10 \tab 2 \tab - \tab 1468 \cr
   1982 \tab - \tab 903 \tab 947 \tab 7 \tab 3 \tab - \tab 1860 \cr
   1984 \tab - \tab 611 \tab 840 \tab 17 \tab 5 \tab - \tab 1473 \cr
   1985 \tab - \tab 617 \tab 901 \tab 13 \tab 3 \tab - \tab 1534 \cr
   1987 \tab - \tab 721 \tab 1083 \tab 6 \tab 9 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 391 \tab 575 \tab 9 \tab 2 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 408 \tab 619 \tab 5 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 374 \tab 530 \tab 10 \tab 3 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 426 \tab 556 \tab 5 \tab 6 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 461 \tab 609 \tab 3 \tab 2 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 933 \tab 1045 \tab 15 \tab 3 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 804 \tab 1099 \tab 20 \tab - \tab - \tab 2904 \cr
   1998 \tab 950 \tab 772 \tab 1079 \tab 28 \tab 3 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 725 \tab 1106 \tab 25 \tab 5 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 296 \tab 623 \tab 5 \tab - \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 283 \tab 612 \tab 3 \tab - \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 719 \tab 1274 \tab 10 \tab - \tab - \tab 4510 \cr
   2008 \tab 671 \tab 453 \tab 896 \tab 3 \tab - \tab - \tab 2023 \cr
   2010 \tab 763 \tab 422 \tab 854 \tab 4 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 445 \tab 855 \tab 7 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 566 \tab 1138 \tab 10 \tab - \tab - \tab 2538 \cr
   2016 \tab 977 \tab 604 \tab 1271 \tab 12 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 521 \tab 1046 \tab 6 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab 1409 \tab 2613 \tab 2 \tab - \tab 8 \tab 4032 \cr
   2022 \tab 1203 \tab 878 \tab 1440 \tab 18 \tab 1 \tab 4 \tab 3544 \cr
   Total \tab 18541 \tab 17988 \tab 27793 \tab 283 \tab 69 \tab 12 \tab 64686 \cr
}
}

\seealso{
Other Core: 
\code{\link{burglr}},
\code{\link{fearhome}},
\code{\link{owngun}},
\code{\link{pistol}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{rifle}},
\code{\link{robbry}},
\code{\link{shotgun}}
}
\concept{Core}
\keyword{variable}
