% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{fearhome}
\alias{fearhome}
\title{Does r feel safe and secure at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
fearhome
}
\details{
Question How about at home at night--do you feel safe and secure, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/676/vshow}{GSS Data Explorer page for \code{fearhome}}.
}

\seealso{
Other Core: 
\code{\link{burglr}},
\code{\link{fear}},
\code{\link{owngun}},
\code{\link{pistol}},
\code{\link{polattak}},
\code{\link{polescap}},
\code{\link{rifle}},
\code{\link{robbry}},
\code{\link{shotgun}}

Other Crime: 
\code{\link{fear}}

Other Residence: 
\code{\link{fear}}
}
\concept{Core}
\concept{Crime}
\concept{Residence}
\keyword{variable}
