% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{febrass}
\alias{febrass}
\title{Do women command over men}
\source{
General Social Survey https://gss.norc.org
}
\description{
febrass
}
\details{
Question C. As far as you know, are women in the armed forces now assigned to jobs where they have command over men, or are women not assigned to such jobs?  RECORD UNDER "C."
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, they are
\item \verb{[2]} no they arent
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/800/vshow}{GSS Data Explorer page for \code{febrass}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, they are \tab no they arent \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 840 \tab 476 \tab 522 \tab 22 \tab 1860 \cr
   Total \tab 0 \tab 840 \tab 476 \tab 522 \tab 22 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{fehlpmil}}
}
\concept{Core}
\keyword{variable}
