% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{fecare}
\alias{fecare}
\title{R's concern about women's rights}
\source{
General Social Survey https://gss.norc.org
}
\description{
fecare
}
\details{
Question A. How concerned are you personally about women''s rights?  Are you very concerned somewhat concerned, not very concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \code{1} very concerned
\item \code{2} somewhat concerned
\item \code{3} not very concerned
\item \code{4} not concerned at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/596/vshow}{GSS Data Explorer page for \code{fecare}}.
}

\seealso{
Other Core: 
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}

Other Split Ballots: 
\code{\link{fefirm}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}}

Other Women: 
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}}
}
\concept{Core}
\concept{Split Ballots}
\concept{Women}
\keyword{variable}
