% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{fechld2}
\alias{fechld2}
\title{Mother working doesn't hurt children - issp version}
\source{
General Social Survey https://gss.norc.org
}
\description{
fechld2
}
\details{
Question To begin, we have some questions about women. To what extent do you agree or disagree...? A working mother can establish just as warm and secure a relationship with her children as a mother who does not work.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7764/vshow}{GSS Data Explorer page for \code{fechld2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2377 \tab 487 \tab 434 \tab 83 \tab 117 \tab 21 \tab 14 \tab 6 \tab 5 \tab 3544 \cr
   Total \tab 2377 \tab 487 \tab 434 \tab 83 \tab 117 \tab 21 \tab 14 \tab 6 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
