% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{feclergy}
\alias{feclergy}
\title{R favors women as ministers}
\source{
General Social Survey https://gss.norc.org
}
\description{
feclergy
}
\details{
Question B. ASK NON-CATHOLICS ONLY.
Do you favor or oppose women as pastors, ministers, priests, or rabbis in your own faith or denomination?
}
\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{8} no opinion
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/380/vshow}{GSS Data Explorer page for \code{feclergy}}.
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}},
\code{\link{relgrade}}
}
\concept{Core}
\keyword{variable}
