% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{fecolop}
\alias{fecolop}
\title{College opportunities for women}
\source{
General Social Survey https://gss.norc.org
}
\description{
fecolop
}
\details{
Question Would you say that opportunities for college education are, in general, better or worse, for women than for men?
}
\section{Values}{

\itemize{
\item \code{1} much bettr for women
\item \code{2} better for women
\item \code{3} no difference
\item \code{4} worse for women
\item \code{5} much wrse for women
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4009/vshow}{GSS Data Explorer page for \code{fecolop}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}},
\code{\link{opfamily}}

Other Opportunity: 
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{opfamily}}

Other Women: 
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{fejobop}}
}
\concept{ISSP Role of Gov't}
\concept{Opportunity}
\concept{Women}
\keyword{variable}
