% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_88.R
\name{fedirty}
\alias{fedirty}
\title{Are women assigned to dirty jobs}
\source{
General Social Survey https://gss.norc.org
}
\description{
fedirty
}
\details{
Question B. As far as you know, are women in the armed forces now assigned to dirty jobs like repairing trucks or other heavy equipment, or are women not assigned to such jobs?
}
\section{Values}{

\itemize{
\item \code{1} yes, they are
\item \code{2} no they arent
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/799/vshow}{GSS Data Explorer page for \code{fedirty}}.
}

\seealso{
Other Core: 
\code{\link{brass}},
\code{\link{fefight}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{gunner}},
\code{\link{mechanic}},
\code{\link{nurse}},
\code{\link{transair}},
\code{\link{typist}}

Other Military: 
\code{\link{brass}},
\code{\link{fefight}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{gunner}},
\code{\link{mechanic}},
\code{\link{nurse}},
\code{\link{transair}},
\code{\link{typist}}
}
\concept{Core}
\concept{Military}
\keyword{variable}
