% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{fedtrust}
\alias{fedtrust}
\title{How much does r trust federal government}
\source{
General Social Survey https://gss.norc.org
}
\description{
fedtrust
}
\details{
Question How much of the time do you think you can trust the government in Washington to do what is right -- just about always, most of the time only some of the time, or almost never?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} most of time
\item \code{3} some of time
\item \code{4} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1038/vshow}{GSS Data Explorer page for \code{fedtrust}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
