% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_300.R
\name{feelasns}
\alias{feelasns}
\title{How r feels towards asian americans}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelasns
}
\details{
Question B. In general, how warm or cool do you feel towards Asian Americans?
}
\section{Values}{

\itemize{
\item \code{1} very warm
\item \code{9} very cool
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2731/vshow}{GSS Data Explorer page for \code{feelasns}}.
}

\seealso{
Other Prejudice: 
\code{\link{ethadapt}},
\code{\link{ethignor}},
\code{\link{ethimp}},
\code{\link{ethnofit}},
\code{\link{ethtrads}},
\code{\link{feelblks}},
\code{\link{feelhsps}}
}
\concept{Asian Americans}
\concept{Prejudice}
\keyword{variable}
