% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{feelnerv}
\alias{feelnerv}
\title{How often felt nervious in past 2 weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelnerv
}
\details{
Question Over the last 2 weeks, how often have you been bothered by the following problems: Feeling nervous, anxious, or on edge.
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} several days
\item \code{3} more than half the days
\item \code{4} nearly every day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7859/vshow}{GSS Data Explorer page for \code{feelnerv}}.
}

\keyword{variable}
