% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{feelrel}
\alias{feelrel}
\title{How religious is r}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelrel
}
\details{
Question Would you describe yourself as. . .
}
\section{Values}{

\itemize{
\item \code{1} extreme relgious
\item \code{2} very religious
\item \code{3} somewhat relgious
\item \code{4} not religious nor non-religious
\item \code{5} somewhat non-religious
\item \code{6} very non-religious
\item \code{7} extremely non-religious
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4679/vshow}{GSS Data Explorer page for \code{feelrel}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab extreme relgious \tab very religious \tab somewhat relgious \tab not religious nor non-religious \tab somewhat non-religious \tab very non-religious \tab extremely non-religious \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 158 \tab 61 \tab 284 \tab 620 \tab 219 \tab 52 \tab 38 \tab 32 \tab 28 \tab 25 \tab 1517 \cr
   1998 \tab 1548 \tab 83 \tab 234 \tab 603 \tab 121 \tab 59 \tab 46 \tab 56 \tab 40 \tab 42 \tab 2832 \cr
   2008 \tab 658 \tab 91 \tab 271 \tab 678 \tab 96 \tab 97 \tab 52 \tab 70 \tab 4 \tab 6 \tab 2023 \cr
   2018 \tab 1173 \tab 73 \tab 249 \tab 451 \tab 136 \tab 82 \tab 92 \tab 73 \tab 8 \tab 11 \tab 2348 \cr
   Total \tab 3537 \tab 308 \tab 1038 \tab 2352 \tab 572 \tab 290 \tab 228 \tab 231 \tab 80 \tab 84 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{prayfreq}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{schlpray}}
}
\concept{ISSP Religion}
\keyword{variable}
