% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{feelrel}
\alias{feelrel}
\title{How religious is r}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelrel
}
\details{
Question Would you describe yourself as. . .
}
\section{Values}{

\itemize{
\item \code{1} extreme relgious
\item \code{2} very religious
\item \code{3} somewhat relgious
\item \code{4} not religious nor non-religious
\item \code{5} somewhat non-religious
\item \code{6} very non-religious
\item \code{7} extremely non-religious
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4679/vshow}{GSS Data Explorer page for \code{feelrel}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{prayfreq}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{schlpray}}

Other Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{prayfreq}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{schlpray}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
