% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{feelwhts}
\alias{feelwhts}
\title{How r feel towards caucasians}
\source{
General Social Survey https://gss.norc.org
}
\description{
feelwhts
}
\details{
Question D. In general, how warm or cool do you feel towards white or Caucasian Americans?
}
\section{Values}{

\itemize{
\item \code{1} very warm
\item \code{9} very cool
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2733/vshow}{GSS Data Explorer page for \code{feelwhts}}.
}

\seealso{
Other Prejudice: 
\code{\link{ethdiff}},
\code{\link{ethsame}},
\code{\link{whtsdiff}}

Other Whites: 
\code{\link{whtsdiff}}
}
\concept{Prejudice}
\concept{Whites}
\keyword{variable}
