% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_606.R
\name{feeused}
\alias{feeused}
\title{Fee given to get case}
\source{
General Social Survey https://gss.norc.org
}
\description{
feeused
}
\details{
Question Was a respondent incentive or fee used as part of getting this case?
}
\section{Values}{

\itemize{
\item \code{1} yes, money
\item \code{2} yes, other (flowers, food, etc.)
\item \code{3} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5560/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab no \tab no answer \tab yes, money \tab yes, other (flowers, food, etc.) \tab don't know \tab not available in this year \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 1253 \tab 54 \tab 1432 \tab 93 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 966 \tab 98 \tab 1695 \tab 58 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 367 \tab 13 \tab 2339 \tab 46 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 1553 \tab 12 \tab 1233 \tab 14 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 2262 \tab 7 \tab 2160 \tab 81 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 804 \tab 3 \tab 1199 \tab 17 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 672 \tab 12 \tab 1336 \tab 20 \tab 4 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 750 \tab 5 \tab 1193 \tab 20 \tab 6 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 925 \tab 5 \tab 1513 \tab 92 \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 603 \tab 11 \tab 2147 \tab 102 \tab 4 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 295 \tab 10 \tab 1953 \tab 5 \tab 85 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab 1642 \tab 135 \tab - \tab 1677 \tab 8 \tab 82 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 36926 \tab 10585 \tab 230 \tab 19877 \tab 556 \tab 184 \tab 4032 \tab 3309 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Paradata: 
\code{\link{feelevel}},
\code{\link{intid}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}

Other Interview: 
\code{\link{feelevel}},
\code{\link{intid}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}}
}
\concept{Interview}
\concept{Paradata}
\keyword{variable}
