% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_88.R
\name{fefight}
\alias{fefight}
\title{Are women assigned to combat}
\source{
General Social Survey https://gss.norc.org
}
\description{
fefight
}
\details{
Question A. As far as you know, are women now assigned to jobs in the armed forces that would expose them to combat, or are women not assigned to such jobs?  RECORD UNDER "A" BELOW.
}
\section{Values}{

\itemize{
\item \code{1} yes, they are
\item \code{2} no they arent
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/798/vshow}{GSS Data Explorer page for \code{fefight}}.
}

\seealso{
Other Core: 
\code{\link{brass}},
\code{\link{fedirty}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{gunner}},
\code{\link{mechanic}},
\code{\link{nurse}},
\code{\link{transair}},
\code{\link{typist}}
}
\concept{Core}
\keyword{variable}
