% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{fegains}
\alias{fegains}
\title{Best way for women to improve position}
\source{
General Social Survey https://gss.norc.org
}
\description{
fegains
}
\details{
Question People have different ideas about the best way for women to improve their position in American life and politics. Some people think that the best way for women to improve their position is through women's rights groups, they would be at point 1 on this card. Other people think that the best way for women to improve their position is for each individual women to become better trained and more qualified. They would be located at point 7. And other people have opinions somewhere in between. Where would you place yourself on this scale?
}
\section{Values}{

\itemize{
\item \code{1} womens rights grps
\item \code{7} better trained inds
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1044/vshow}{GSS Data Explorer page for \code{fegains}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab womens rights grps \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab better trained inds \tab no answer \tab Total \cr
   1987 \tab 778 \tab 59 \tab 27 \tab 38 \tab 188 \tab 136 \tab 130 \tab 442 \tab 21 \tab 1819 \cr
   Total \tab 778 \tab 59 \tab 27 \tab 38 \tab 188 \tab 136 \tab 130 \tab 442 \tab 21 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{blkgains}},
\code{\link{blkinflu}},
\code{\link{feinflu}},
\code{\link{gavepol}}
}
\concept{Socio-political participation}
\keyword{variable}
