% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{fehire}
\alias{fehire}
\title{Should hire and promote women}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehire
}
\details{
Question Now I'm going to read several statements. As I read each one, please tell me whether you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree. For example, here is the statement:
Because of past discrimination, employers should make special efforts to hire and promote qualified women.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1697/vshow}{GSS Data Explorer page for \code{fehire}}.
}

\seealso{
Other Gender: 
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{hmemaker}},
\code{\link{mebear}},
\code{\link{wrkclass}}

Other Women: 
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{hmemaker}},
\code{\link{mebear}},
\code{\link{wrkclass}}
}
\concept{Affirmative Action}
\concept{Gender}
\concept{Job}
\concept{Women}
\keyword{variable}
