% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{fehlpbus}
\alias{fehlpbus}
\title{Govt should aid women seeking jobs}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehlpbus
}
\details{
Question Here are three things the government might do. Some people are in favor of them while other people are against them. Please circle one number for each statement to show how you feel.
A. The government should increase opportunities for women in business and industry.
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor
\item \code{2} in favor
\item \code{3} neither
\item \code{4} against
\item \code{5} strongly against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4012/vshow}{GSS Data Explorer page for \code{fehlpbus}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fecolop}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}},
\code{\link{opfamily}}

Other Opportunity: 
\code{\link{fecolop}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{opfamily}}

Other Women: 
\code{\link{fecolop}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{fejobop}}
}
\concept{ISSP Role of Gov't}
\concept{Opportunity}
\concept{Women}
\keyword{variable}
