% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{fehlpcol}
\alias{fehlpcol}
\title{Govt should aid women seeking higher ed}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehlpcol
}
\details{
Question Here are three things the government might do. Some people are in favor of them while other people are against them. Please circle one number for each statement to show how you feel.
B. The government should increase opportunities for women to go to college.
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor
\item \code{2} in favor
\item \code{3} neither
\item \code{4} against
\item \code{5} strongly against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4013/vshow}{GSS Data Explorer page for \code{fehlpcol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly in favor \tab in favor \tab neither \tab against \tab strongly against \tab no answer \tab Total \cr
   1985 \tab 857 \tab 86 \tab 189 \tab 254 \tab 110 \tab 15 \tab 23 \tab 1534 \cr
   Total \tab 857 \tab 86 \tab 189 \tab 254 \tab 110 \tab 15 \tab 23 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}},
\code{\link{opfamily}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
