% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_89.R
\name{fehlpmil}
\alias{fehlpmil}
\title{Do women raise effectiveness of military}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehlpmil
}
\details{
Question In the past 10 years the number of women in the armed forces has increased from about 2 percent to about 8 percent
today. In general, would you say the increased number of women has raised the effectiveness of our armed forces, has it made no difference, or has it made our armed forces less effective?
(Just your own opinion.)
}
\section{Values}{

\itemize{
\item \code{1} raised effectiveness
\item \code{2} no difference
\item \code{3} lowered effctiveness
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/801/vshow}{GSS Data Explorer page for \code{fehlpmil}}.
}

\seealso{
Other Core: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}}

Other Military: 
\code{\link{draft}},
\code{\link{draftcol}},
\code{\link{draftem}},
\code{\link{draftfe}},
\code{\link{draftfem}},
\code{\link{draftgay}},
\code{\link{draftmar}},
\code{\link{draftpar}},
\code{\link{febrass}}
}
\concept{Core}
\concept{Military}
\keyword{variable}
