% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{fehome}
\alias{fehome}
\title{Women take care of home not country}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehome
}
\details{
Question Do you agree or disagree with this statement?  Women should take care of running their homes and leave running the country up to men.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/588/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 509 \tab 922 \tab 53 \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 515 \tab 931 \tab 42 \tab 2 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 569 \tab 921 \tab 38 \tab 2 \tab - \tab 1530 \cr
   1978 \tab - \tab 473 \tab 1009 \tab 48 \tab 2 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 511 \tab 1298 \tab 50 \tab 1 \tab - \tab 1860 \cr
   1983 \tab - \tab 357 \tab 1182 \tab 55 \tab 5 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 394 \tab 1101 \tab 34 \tab 5 \tab - \tab 1534 \cr
   1986 \tab - \tab 344 \tab 1080 \tab 43 \tab 3 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 201 \tab 749 \tab 34 \tab 4 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 192 \tab 767 \tab 43 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 159 \tab 731 \tab 32 \tab 6 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 190 \tab 793 \tab 35 \tab 6 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 154 \tab 888 \tab 34 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 268 \tab 1632 \tab 71 \tab 6 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 308 \tab 1576 \tab 69 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 280 \tab 1534 \tab 50 \tab 7 \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 45513 \tab 5424 \tab 17114 \tab 731 \tab 64 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}

Other Women: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}}
}
\concept{Core}
\concept{Women}
\keyword{variable}
