% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{feimp}
\alias{feimp}
\title{Importance of women's rights issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
feimp
}
\details{
Question A. How important is the women''s rights issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \code{1} one of most imp
\item \code{2} important
\item \code{3} not very imp
\item \code{4} not imp at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/593/vshow}{GSS Data Explorer page for \code{feimp}}.
}

\seealso{
Other Core: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}

Other Split Ballots: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}}

Other Women: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepoly}},
\code{\link{fepres}},
\code{\link{fework}}
}
\concept{Core}
\concept{Gender}
\concept{Split Ballots}
\concept{Women}
\keyword{variable}
