% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{feinc}
\alias{feinc}
\title{Income for women vs. men with same bkgd}
\source{
General Social Survey https://gss.norc.org
}
\description{
feinc
}
\details{
Question And how about income and wages -- compared with men who have similar education and jobs -- are women In general paid better or worse than men?
}
\section{Values}{

\itemize{
\item \verb{[1]} much bettr for women
\item \verb{[2]} better for women
\item \verb{[3]} no difference
\item \verb{[4]} worse for women
\item \verb{[5]} much wrse for women
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4011/vshow}{GSS Data Explorer page for \code{feinc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much bettr for women \tab better for women \tab no difference \tab worse for women \tab much wrse for women \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 7 \tab 8 \tab 135 \tab 432 \tab 49 \tab 35 \tab 11 \tab 1534 \cr
   Total \tab 857 \tab 7 \tab 8 \tab 135 \tab 432 \tab 49 \tab 35 \tab 11 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}},
\code{\link{opfamily}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
