% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{fejobaff}
\alias{fejobaff}
\title{For or against preferential hiring of women}
\source{
General Social Survey https://gss.norc.org
}
\description{
fejobaff
}
\details{
Question Some people say that because of past discrimination, women should be given preference in hiring and promotion. Others say that such preference in hiring and promotion of women is wrong because it discriminates against men. What about your opinion - are you for or against preferential hiring and promotion of women?
IF FOR:Do you favor preference in hiring and promotion strongly or not strongly?
IF AGAINST:Do you oppose preference in hiring and promotion strongly or not strongly?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} not strongly favor
\item \code{3} not strongly oppose
\item \code{4} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
   2004 \tab A/B \tab partial \cr
   2006 \tab A/B \tab partial \cr
   2008 \tab A/B \tab partial \cr
   2010 \tab A/B \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1656/vshow}{GSS Data Explorer page for \code{fejobaff}}.
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{allhrt}},
\code{\link{balwkfam}},
\code{\link{children}},
\code{\link{chldben}},
\code{\link{chldhrt}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{earnmore}},
\code{\link{emoteoth}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{febear}},
\code{\link{fehire}},
\code{\link{feimp}},
\code{\link{fekids1}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{flextime}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{haprom}},
\code{\link{hmemaker}},
\code{\link{livewith}},
\code{\link{manprof}},
\code{\link{marfree}},
\code{\link{mebear}},
\code{\link{men}},
\code{\link{menben}},
\code{\link{menhrt}},
\code{\link{meovrwrk}},
\code{\link{noathome}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{parleave}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{rhmewrk}},
\code{\link{romance}},
\code{\link{rushed}},
\code{\link{selffrst}},
\code{\link{sexhar}},
\code{\link{sexpromo}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{twoincs}},
\code{\link{willwed1}},
\code{\link{willwed2}},
\code{\link{womenben}},
\code{\link{womenhrt}},
\code{\link{workless}},
\code{\link{worklife}},
\code{\link{wrkclass}},
\code{\link{yourself}}
}
\concept{Gender}
\keyword{variable}
