% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_180.R
\name{fekids1}
\alias{fekids1}
\title{How important women biologically better-suited}
\source{
General Social Survey https://gss.norc.org
}
\description{
fekids1
}
\details{
Question Women are more likely than men to take care of children. I'm going to read several reasons why this might be so. please tell me how important you think each reason is: very important, important, somewhat important, or not at all important. For example, ...
A. Women are biologically better-suited to care for children. How important do you think this reason is for explaining why women are more likely to take care of children?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} somewhat important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1691/vshow}{GSS Data Explorer page for \code{fekids1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 480 \tab 371 \tab 325 \tab 258 \tab 22 \tab 4 \tab 2904 \cr
   Total \tab 1444 \tab 480 \tab 371 \tab 325 \tab 258 \tab 22 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{feless1}},
\code{\link{feless2}},
\code{\link{feless3}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{womenhrt}}
}
\concept{Gender}
\keyword{variable}
