% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{fekids2}
\alias{fekids2}
\title{How important women are taught from childhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
fekids2
}
\details{
Question Women are more likely than men to take care of children. I'm going to read several reasons why this might be so. please tell me how important you think each reason is: very important, important, somewhat important, or not at all important. For example, ...
B. Women are taught from childhood how to care for children. How important do you think this reason is for explaining why women are more likely to take care of children?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} somewhat important
\item \verb{[4]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1692/vshow}{GSS Data Explorer page for \code{fekids2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 463 \tab 445 \tab 319 \tab 213 \tab 16 \tab 4 \tab 2904 \cr
   Total \tab 1444 \tab 463 \tab 445 \tab 319 \tab 213 \tab 16 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{fehire}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feminist}},
\code{\link{fenews}},
\code{\link{hmemaker}},
\code{\link{mebear}},
\code{\link{wrkclass}}
}
\concept{Gender}
\keyword{variable}
