% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_180.R
\name{feless2}
\alias{feless2}
\title{How important women's family responsibilities}
\source{
General Social Survey https://gss.norc.org
}
\description{
feless2
}
\details{
Question On average, women who are employed full time earn less than men earn. I'm going to read several reasons why this might be so. Please tell me how important you think each reason is: very important, important, somewhat important, or not at all important. For example,...
B. Women's family responsibilities keep them from putting as much time and effort into their jobs as men do. How important Do you think this reason is for explaining why women earn less?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} somewhat important
\item \code{4} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1689/vshow}{GSS Data Explorer page for \code{feless2}}.
}

\seealso{
Other Gender: 
\code{\link{allhrt}},
\code{\link{chldhrt}},
\code{\link{fekids1}},
\code{\link{feless1}},
\code{\link{feless3}},
\code{\link{menhrt}},
\code{\link{nooneben}},
\code{\link{noonehrt}},
\code{\link{womenhrt}}

Other Employment: 
\code{\link{feless1}},
\code{\link{feless3}}

Other Job: 
\code{\link{feless1}},
\code{\link{feless3}}

Other Women: 
\code{\link{fekids1}},
\code{\link{feless1}},
\code{\link{feless3}},
\code{\link{womenhrt}}

Other Work: 
\code{\link{feless1}},
\code{\link{feless3}}
}
\concept{Employment}
\concept{Gender}
\concept{Job}
\concept{Women}
\concept{Work}
\keyword{variable}
