% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{femarry}
\alias{femarry}
\title{Way of life recommendation for woman}
\source{
General Social Survey https://gss.norc.org
}
\description{
femarry
}
\details{
Question If you were advising a young woman, which of the following ways of life would you recommend?
}
\section{Values}{

\itemize{
\item \code{1} live alone
\item \code{2} live with posslq
\item \code{3} cohabitate, then marry
\item \code{4} marry
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4362/vshow}{GSS Data Explorer page for \code{femarry}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
