% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{femilop}
\alias{femilop}
\title{Fem opportunities better in military?}
\source{
General Social Survey https://gss.norc.org
}
\description{
femilop
}
\details{
Question B. And what about women?  Would you say that their treatment and opportunities are better in the military, better in civilian employment, or that there isn''t any difference these days?
}
\section{Values}{

\itemize{
\item \code{1} military better
\item \code{2} civilian better
\item \code{3} no difference
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/815/vshow}{GSS Data Explorer page for \code{femilop}}.
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}

Other Military: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfamnw}}
}
\concept{Core}
\concept{Military}
\keyword{variable}
