% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{feminist}
\alias{feminist}
\title{Think yourself as a feminist?}
\source{
General Social Survey https://gss.norc.org
}
\description{
feminist
}
\details{
Question Do you think of yourself as a feminist or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, a feminist
\item \verb{[2]} no, not a feminist
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1698/vshow}{GSS Data Explorer page for \code{feminist}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, a feminist \tab no, not a feminist \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 298 \tab 1083 \tab 77 \tab 2 \tab 2904 \cr
   Total \tab 1444 \tab 298 \tab 1083 \tab 77 \tab 2 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{fehire}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{fenews}},
\code{\link{hmemaker}},
\code{\link{mebear}},
\code{\link{wrkclass}}
}
\concept{Gender}
\keyword{variable}
