% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{femsee1}
\alias{femsee1}
\title{How feminine others see r (0 to 6 scale)}
\source{
General Social Survey https://gss.norc.org
}
\description{
femsee1
}
\details{
Question On that same scale, how feminine do most people see you?

NOTE: FEMSELF1, FEMSEE1, MASCSELF1 and MASCSEE1 are part of a BALLOT/FORM randomization. On BALLOT=2, all items were asked at the start of the High Risk Behaviors module, and on BALLOT=3, all items were asked at the end. For FORM=1, FEMSELF1 and FEMSEE1 were asked first, and for FORM=2, MASCSELF1 and MASCSEE1 were asked first.
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{6} very
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8187/vshow}{GSS Data Explorer page for \code{femsee1}}.
}

\seealso{
Other Core: 
\code{\link{family16sex}},
\code{\link{femself1}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{racdif5x}},
\code{\link{sexnow2}}
}
\concept{Core}
\keyword{variable}
