% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{fenews}
\alias{fenews}
\title{How often pay attention to women issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
fenews
}
\details{
Question When reading, watching, or listening to the news, how often do you pay attention to issues that especially affect women - very often, fairly often, not too often, or never?
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} fairly often
\item \code{3} not too often
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1699/vshow}{GSS Data Explorer page for \code{fenews}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab fairly often \tab not too often \tab never \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 325 \tab 601 \tab 461 \tab 66 \tab 3 \tab 4 \tab 2904 \cr
   Total \tab 1444 \tab 325 \tab 601 \tab 461 \tab 66 \tab 3 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{fehire}},
\code{\link{fekids2}},
\code{\link{fekids3}},
\code{\link{fekids4}},
\code{\link{fekids5}},
\code{\link{feminist}},
\code{\link{hmemaker}},
\code{\link{mebear}},
\code{\link{wrkclass}}
}
\concept{Gender}
\keyword{variable}
