% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{fenumok}
\alias{fenumok}
\title{Number of women in military}
\source{
General Social Survey https://gss.norc.org
}
\description{
fenumok
}
\details{
Question As you know, this country stopped the military draft in 1972. Since that time we have relied on volunteers. Now I'd like to ask you a few questions about our armed forces.
C. At the present time, about 9 percent of the armed forces are women. All things considered, do you think there are too many women in the armed forces, about the right number, or should there be more women in the armed forces?
RECORD UNDER "C" BELOW.
}
\section{Values}{

\itemize{
\item \verb{[1]} too many
\item \verb{[2]} about right number
\item \verb{[3]} should be more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/783/vshow}{GSS Data Explorer page for \code{fenumok}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too many \tab about right number \tab should be more \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 158 \tab 965 \tab 550 \tab 173 \tab 14 \tab 1860 \cr
   1983 \tab - \tab 120 \tab 864 \tab 476 \tab 133 \tab 6 \tab 1599 \cr
   1984 \tab - \tab 108 \tab 785 \tab 460 \tab 93 \tab 27 \tab 1473 \cr
   Total \tab 0 \tab 386 \tab 2614 \tab 1486 \tab 399 \tab 47 \tab 4932 \cr
}
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
