% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{fepolv}
\alias{fepolv}
\title{Women not suited for politics (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fepolv
}
\details{
Question Please indicate if you agree or disagree with this statement: Most men are better suited emotionally for politics than are most women.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{3} not sure (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7602/vshow}{GSS Data Explorer page for \code{fepolv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2658 \tab 126 \tab 1104 \tab 140 \tab 4 \tab 4032 \cr
   2022 \tab 2988 \tab 51 \tab 437 \tab 68 \tab - \tab 3544 \cr
   Total \tab 5646 \tab 177 \tab 1541 \tab 208 \tab 4 \tab 7576 \cr
}
}

\keyword{variable}
