% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{fepoly}
\alias{fepoly}
\title{Men or women better suited for politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
fepoly
}
\details{
Question B. Would you say that most men are better suited emotionally for politics than are most women, that men and women are equally suited, or that women are better suited than men in this area?
}
\section{Values}{

\itemize{
\item \code{1} men
\item \code{2} equal
\item \code{3} women
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/592/vshow}{GSS Data Explorer page for \code{fepoly}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab men \tab equal \tab women \tab don't know \tab no answer \tab Total \cr
   1974 \tab 753 \tab 238 \tab 450 \tab 31 \tab 11 \tab 1 \tab 1484 \cr
   1982 \tab 923 \tab 229 \tab 630 \tab 50 \tab 21 \tab 7 \tab 1860 \cr
   Total \tab 1676 \tab 467 \tab 1080 \tab 81 \tab 32 \tab 8 \tab 3344 \cr
}
}

\seealso{
Other Core: 
\code{\link{fecare}},
\code{\link{fefirm}},
\code{\link{fehome}},
\code{\link{feimp}},
\code{\link{feinfo}},
\code{\link{fepol}},
\code{\link{fepres}},
\code{\link{fework}},
\code{\link{kidssol}}
}
\concept{Core}
\keyword{variable}
