% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{fepriest}
\alias{fepriest}
\title{R favors women as priests}
\source{
General Social Survey https://gss.norc.org
}
\description{
fepriest
}
\details{
Question A. IF CATHOLIC RESPONDENT, ASK FEPRIEST. OTHERS GO TO FECLERGY.
As I read this statement, will you tell me whether you agree strongly, agree somewhat, disagree somewhat, or disagree strongly?
It would be a good thing if women were allowed to be ordained as priests.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} disagree strongly
\item \code{8} no opinion
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/379/vshow}{GSS Data Explorer page for \code{fepriest}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1986 \tab 1088 \tab 88 \tab 111 \tab 58 \tab 105 \tab 14 \tab 6 \tab 1470 \cr
   Total \tab 1088 \tab 88 \tab 111 \tab 58 \tab 105 \tab 14 \tab 6 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}},
\code{\link{relgrade}}
}
\concept{Core}
\keyword{variable}
