% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{feworkif}
\alias{feworkif}
\title{Married women work if jobs limited}
\source{
General Social Survey https://gss.norc.org
}
\description{
feworkif
}
\details{
Question If there is a limited number of jobs, do you approve or disapprove of a married woman holding a job in business or industry when her husband is able to support her?
}
\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/718/vshow}{GSS Data Explorer page for \code{feworkif}}.
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{racsubgv}},
\code{\link{racsubs}}

Other Women: 
\code{\link{febear}}
}
\concept{Core}
\concept{Women}
\keyword{variable}
