% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{fewrksup}
\alias{fewrksup}
\title{Do most women work to support family}
\source{
General Social Survey https://gss.norc.org
}
\description{
fewrksup
}
\details{
Question Do you agree or disagree . . . Most women have to work these days to support their families.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4401/vshow}{GSS Data Explorer page for \code{fewrksup}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 403 \tab 800 \tab 122 \tab 77 \tab 21 \tab 9 \tab 15 \tab 2992 \cr
   Total \tab 1545 \tab 403 \tab 800 \tab 122 \tab 77 \tab 21 \tab 9 \tab 15 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabfst}},
\code{\link{cohabok}},
\code{\link{divbest}},
\code{\link{divifkd1}},
\code{\link{divnokd1}},
\code{\link{hubbywk1}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{singlpar}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
