% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{fightair}
\alias{fightair}
\title{Women as jet fighter pilots}
\source{
General Social Survey https://gss.norc.org
}
\description{
fightair
}
\details{
Question I'm going to read you a list of jobs that people might have in the armed forces. Please tell me whether you think a woman should or should not be assigned to each job, assuming she is trained to do it.
A. A jet fighter pilot ...
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/789/vshow}{GSS Data Explorer page for \code{fightair}}.
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
