% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{findjob}
\alias{findjob}
\title{Could r find a job (unemployed r's)}
\source{
General Social Survey https://gss.norc.org
}
\description{
findjob
}
\details{
Question If you were looking actively, how easy or difficult do you think it would be for you to find an acceptable job?
}
\section{Values}{

\itemize{
\item \verb{[1]} very easy
\item \verb{[2]} fairly easy
\item \verb{[3]} neither easy nor difficult
\item \verb{[4]} fairly difficult
\item \verb{[5]} very difficult
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4530/vshow}{GSS Data Explorer page for \code{findjob}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very easy \tab fairly easy \tab neither easy nor difficult \tab fairly difficult \tab very difficult \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 1253 \tab 13 \tab 44 \tab 22 \tab 45 \tab 41 \tab 10 \tab 109 \tab 1537 \cr
   1998 \tab 1609 \tab 234 \tab 365 \tab 164 \tab 236 \tab 148 \tab 44 \tab 32 \tab 2832 \cr
   Total \tab 2862 \tab 247 \tab 409 \tab 186 \tab 281 \tab 189 \tab 54 \tab 141 \tab 4369 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
