% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{findjob}
\alias{findjob}
\title{Could r find a job (unemployed r's)}
\source{
General Social Survey https://gss.norc.org
}
\description{
findjob
}
\details{
Question If you were looking actively, how easy or difficult do you think it would be for you to find an acceptable job?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} fairly easy
\item \code{3} neither easy nor difficult
\item \code{4} fairly difficult
\item \code{5} very difficult
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4530/vshow}{GSS Data Explorer page for \code{findjob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Employment}
\concept{Work Orientation}
\keyword{variable}
