% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{findnwjb}
\alias{findnwjb}
\title{How likely r to find job with same/better benefit}
\source{
General Social Survey https://gss.norc.org
}
\description{
findnwjb
}
\details{
Question If you were laid off, how easy do you think it would be for you to find a job with another employer with approximately the same or better retirement and health benefits you now have? Would you say…
}
\section{Values}{

\itemize{
\item \verb{[1]} very easy
\item \verb{[2]} somewhat easy
\item \verb{[3]} not at all easy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3883/vshow}{GSS Data Explorer page for \code{findnwjb}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very easy \tab somewhat easy \tab not at all easy \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 181 \tab 280 \tab 370 \tab 14 \tab 13 \tab 2023 \cr
   Total \tab 1165 \tab 181 \tab 280 \tab 370 \tab 14 \tab 13 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\keyword{variable}
