% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{findnwjb}
\alias{findnwjb}
\title{How likely r to find job with same/better benefit}
\source{
General Social Survey https://gss.norc.org
}
\description{
findnwjb
}
\details{
Question If you were laid off, how easy do you think it would be for you to find a job with another employer with approximately the same or better retirement and health benefits you now have? Would you say…
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} somewhat easy
\item \code{3} not at all easy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3883/vshow}{GSS Data Explorer page for \code{findnwjb}}.
}

\seealso{
Other Global Economics: 
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}

Other Job: 
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\concept{Health}
\concept{Job}
\concept{Occupation}
\keyword{variable}
