% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{findwork}
\alias{findwork}
\title{Could r find a job (employed r's)}
\source{
General Social Survey https://gss.norc.org
}
\description{
findwork
}
\details{
Question If you lost your job for any reason, and were looking actively for another one, how easy or difficult do you think it would be for you to find an acceptable job?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} fairly easy
\item \code{3} neither easy nor difficult
\item \code{4} fairly difficult
\item \code{5} very difficult
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4558/vshow}{GSS Data Explorer page for \code{findwork}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very easy \tab fairly easy \tab neither easy nor difficult \tab fairly difficult \tab very difficult \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 182 \tab 321 \tab 138 \tab 133 \tab 76 \tab 24 \tab 24 \tab 1537 \cr
   Total \tab 639 \tab 182 \tab 321 \tab 138 \tab 133 \tab 76 \tab 24 \tab 24 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
