% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{fininc2}
\alias{fininc2}
\title{Best time to be told about incentives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fininc2
}
\details{
Question Please assume that your health plan has a financial incentive for your doctor to limit the use of expensive tests.
OR
Please assume that you have a health plan and the plan has a financial incentive for your doctor to limit the use of expensive tests
When do you think is the best time to be told about these financial incentives?
}
\section{Values}{

\itemize{
\item \verb{[1]} when you enroll in a health plan
\item \verb{[2]} when you first see your doctor
\item \verb{[3]} when you need a test or specialist
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2749/vshow}{GSS Data Explorer page for \code{fininc2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab when you enroll in a health plan \tab when you first see your doctor \tab when you need a test or specialist \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 2183 \tab 353 \tab 177 \tab 44 \tab 8 \tab 2765 \cr
   Total \tab 0 \tab 2183 \tab 353 \tab 177 \tab 44 \tab 8 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
