% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{fininc3}
\alias{fininc3}
\title{Best person to tell about incentives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fininc3
}
\details{
Question Please assume that your health plan has a financial incentive for your doctor to limit the use of expensive tests.
OR
Please assume that you have a health plan and the plan has a financial incentive for your doctor to limit the use of expensive tests
Who is the best person to tell you about these financial incentives?
}
\section{Values}{

\itemize{
\item \code{1} your doctor
\item \code{2} a representative of your health plan
\item \code{3} both your doctor and a representative of health plan
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2750/vshow}{GSS Data Explorer page for \code{fininc3}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{docaskme}},
\code{\link{doccosts}},
\code{\link{docdecid}},
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docrely}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{docvig}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc4}},
\code{\link{hlthplan}},
\code{\link{insrchng}},
\code{\link{insrlmts}},
\code{\link{regdoc}},
\code{\link{regdocyr}},
\code{\link{secopin}}
}
\concept{Doctors & Patients}
\keyword{variable}
