% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{fininc4}
\alias{fininc4}
\title{R likely to ask doc about incentives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fininc4
}
\details{
Question Please assume that your health plan has a financial incentive for your doctor to limit the use of expensive tests.
OR
Please assume that you have a health plan and the plan has a financial incentive for your doctor to limit the use of expensive tests
How likely would you be to ask your regular doctor whether he or she has financial incentives that limit the use of expensive tests?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} neither likely nor unlikely
\item \code{4} somewhat unlikely
\item \code{5} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2751/vshow}{GSS Data Explorer page for \code{fininc4}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}

Other Health: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{hlthplan}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\concept{Health}
\keyword{variable}
