% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{finind1}
\alias{finind1}
\title{Aged should be financially independent}
\source{
General Social Survey https://gss.norc.org
}
\description{
finind1
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important  is it for them to be...
Financially independent from their parents/guardians
A. By what age should this normally occur?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2897/vshow}{GSS Data Explorer page for \code{finind1}}.
}

\seealso{
Other Adult Transitions: 
\code{\link{eddone}},
\code{\link{eddone1}},
\code{\link{finind}},
\code{\link{ftwork}},
\code{\link{ftwork1}},
\code{\link{ownhh}},
\code{\link{ownhh1}}

Other Income: 
\code{\link{finind}}
}
\concept{Adult Transitions}
\concept{Income}
\keyword{variable}
