% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{firstkidsex}
\alias{firstkidsex}
\title{Sex of first child}
\source{
General Social Survey https://gss.norc.org
}
\description{
firstkidsex
}
\details{
Question Was your first biological child's sex recorded as male or female at birth?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7958/vshow}{GSS Data Explorer page for \code{firstkidsex}}.
}

\seealso{
Other GSS Next: 
\code{\link{femself}},
\code{\link{mascself}},
\code{\link{nonbinkids}}
}
\concept{GSS Next}
\keyword{variable}
