% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{firstslf}
\alias{firstslf}
\title{First job self-employment?}
\source{
General Social Survey https://gss.norc.org
}
\description{
firstslf
}
\details{
Question And how about your first job -- the first job you had after you finished full-time education?
A. Were you self-employed, or did you work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[1]} self-employed
\item \verb{[2]} someone else
\item \verb{[3]} never had job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4317/vshow}{GSS Data Explorer page for \code{firstslf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab self-employed \tab someone else \tab never had job \tab no answer \tab Total \cr
   1987 \tab 255 \tab 57 \tab 1252 \tab 55 \tab 200 \tab 1819 \cr
   Total \tab 255 \tab 57 \tab 1252 \tab 55 \tab 200 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
