% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{firstyou}
\alias{firstyou}
\title{Should take care of oneself and ones family first}
\source{
General Social Survey https://gss.norc.org
}
\description{
firstyou
}
\details{
Question Please select a response to indicate how much you agree or disagree with each of the following statements.
B. You should take care of yourself and your family first, before helping other people
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4230/vshow}{GSS Data Explorer page for \code{firstyou}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 494 \tab 455 \tab 146 \tab 41 \tab 3 \tab 1 \tab 9 \tab 2765 \cr
   2018 \tab 1175 \tab 464 \tab 501 \tab 131 \tab 58 \tab 10 \tab 6 \tab 3 \tab 2348 \cr
   Total \tab 2791 \tab 958 \tab 956 \tab 277 \tab 99 \tab 13 \tab 7 \tab 12 \tab 5113 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{helpjob}},
\code{\link{kidpars}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{talkedto}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
