% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{fishexp2}
\alias{fishexp2}
\title{Why did you choose that answer}
\source{
General Social Survey https://gss.norc.org
}
\description{
fishexp2
}
\details{
Question Why did you choose that answer?
}
\section{Values}{

\itemize{
\item \code{1} because i already know what affects the behavior of fish
\item \code{2} because that is what is allowed to change in this experiment
\item \code{3} because that is what stays the same in this experiment
\item \code{4} because that is what the scientist decided to include in this experiment
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3486/vshow}{GSS Data Explorer page for \code{fishexp2}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}

Other Knowledge: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}

Other Science: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
