% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{fixer1}
\alias{fixer1}
\title{What was the most important reason?}
\source{
General Social Survey https://gss.norc.org
}
\description{
fixer1
}
\details{
Question Think about the most recent time you've paid for home maintenance service -- e.g. a roofer, a painter, a plumber or something like that. What made you choose the person or business that did the work? (Choose all that apply.)
G. Most important reason from above
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1846/vshow}{GSS Data Explorer page for \code{fixer1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2721 \tab 44 \tab 27 \tab 10 \tab 31 \tab 6 \tab 53 \tab 2 \tab 10 \tab 2904 \cr
   Total \tab 2721 \tab 44 \tab 27 \tab 10 \tab 31 \tab 6 \tab 53 \tab 2 \tab 10 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{relfixer}},
\code{\link{satfixes}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
