% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{fixer1}
\alias{fixer1}
\title{What was the most important reason?}
\source{
General Social Survey https://gss.norc.org
}
\description{
fixer1
}
\details{
Question Think about the most recent time you've paid for home maintenance service -- e.g. a roofer, a painter, a plumber or something like that. What made you choose the person or business that did the work? (Choose all that apply.)
G. Most important reason from above
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1846/vshow}{GSS Data Explorer page for \code{fixer1}}.
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{relfixer}},
\code{\link{satfixes}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
