% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{fixera}
\alias{fixera}
\title{My own previous knowledge of the firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
fixera
}
\details{
Question Think about the most recent time you've paid for home maintenance service -- e.g. a roofer, a painter, a plumber or something like that. What made you choose the person or business that did the work? (Choose all that apply.)
A. My own previous knowledge of the firm or workers.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1840/vshow}{GSS Data Explorer page for \code{fixera}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2353 \tab 169 \tab 270 \tab 1 \tab 111 \tab 2904 \cr
   Total \tab 2353 \tab 169 \tab 270 \tab 1 \tab 111 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{rellaw}},
\code{\link{satlawyr}}
}
\concept{Markets}
\keyword{variable}
