% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{fixhome}
\alias{fixhome}
\title{Home maintenance services past 10 years?}
\source{
General Social Survey https://gss.norc.org
}
\description{
fixhome
}
\details{
Question Have you purchased home maintenance services (painting, roofing, plumbing, or something similar) in the past 10 years?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1839/vshow}{GSS Data Explorer page for \code{fixhome}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   1996 \tab 1460 \tab 551 \tab 893 \tab 2904 \cr
   Total \tab 1460 \tab 551 \tab 893 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{rellaw}},
\code{\link{satlawyr}}
}
\concept{Markets}
\keyword{variable}
