% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{fixmistk}
\alias{fixmistk}
\title{How likely govt admnstrtor's be corrected when making}
\source{
General Social Survey https://gss.norc.org
}
\description{
fixmistk
}
\details{
Question When the government administrators make serious mistakes in America how likely is it that they will be corrected?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4958/vshow}{GSS Data Explorer page for \code{fixmistk}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{gvtrghts}},
\code{\link{polinter}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
