% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{flexhrs1}
\alias{flexhrs1}
\title{Importance of flexible hrs. at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
flexhrs1
}
\details{
Question (HAND CARD D2)
(For each of the following, please tell me how important you personally think it is in a job. How important is...)
...a job that allows someone to decide their times or days of work?
(Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} neither important nor unimportant
\item \verb{[4]} not important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6321/vshow}{GSS Data Explorer page for \code{flexhrs1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab neither important nor unimportant \tab not important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 245 \tab 519 \tab 452 \tab 225 \tab 23 \tab 5 \tab 8 \tab 2867 \cr
   Total \tab 1390 \tab 245 \tab 519 \tab 452 \tab 225 \tab 23 \tab 5 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
