% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{fndjob2}
\alias{fndjob2}
\title{Heard job from relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
fndjob2
}
\details{
Question There are many ways people hear about jobs -- such as from ads, employment agencies, or from other people. Could you tell me how you found out about work at your present employer (IF SELF-EMPLOYED:  current organization)?  Please tell me all the ways that you found out about this job.
B. From one of my relatives.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1277/vshow}{GSS Data Explorer page for \code{fndjob2}}.
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{afailure}},
\code{\link{automatn}},
\code{\link{baseofop}},
\code{\link{chngeorg}},
\code{\link{cowrkrel}},
\code{\link{drugtest}},
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{fndjob1}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{fndjob9}},
\code{\link{fringe1}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{futpromo}},
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{genehire}},
\code{\link{genejob}},
\code{\link{helporg}},
\code{\link{idecide}},
\code{\link{imatter}},
\code{\link{imppromo}},
\code{\link{intltest}},
\code{\link{jobcntrl}},
\code{\link{jobjudge}},
\code{\link{jobvshme}},
\code{\link{jobyears}},
\code{\link{learning}},
\code{\link{levels}()},
\code{\link{lottosay}},
\code{\link{ltrsref}},
\code{\link{merged}},
\code{\link{monitred}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{numpromo}},
\code{\link{numunemp}},
\code{\link{offsup}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{orgfin}},
\code{\link{orgmoney}},
\code{\link{othduty1}},
\code{\link{othduty2}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{othpay}},
\code{\link{outofbiz}},
\code{\link{pessimst}},
\code{\link{physical}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{promoted}},
\code{\link{promotng}},
\code{\link{proudorg}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{racpromo}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}},
\code{\link{reorg}},
\code{\link{samejob}},
\code{\link{samevals}},
\code{\link{satself}},
\code{\link{schoolng}},
\code{\link{sexpromo}},
\code{\link{skiltest}},
\code{\link{slfrspct}},
\code{\link{splocnum}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{supduty1}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}},
\code{\link{supothrs}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}},
\code{\link{totmoney}},
\code{\link{totsup}},
\code{\link{totunemp}},
\code{\link{training}},
\code{\link{unmanrel}},
\code{\link{unpower}},
\code{\link{unprog}},
\code{\link{unvote}},
\code{\link{wherewrk}},
\code{\link{wrkindep}},
\code{\link{wrkmuch}},
\code{\link{wrkwell}},
\code{\link{wrkyears}},
\code{\link{youmoney}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
