% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{fndjob4}
\alias{fndjob4}
\title{Heard job from acquaintances}
\source{
General Social Survey https://gss.norc.org
}
\description{
fndjob4
}
\details{
Question There are many ways people hear about jobs -- such as from ads, employment agencies, or from other people. Could you tell me how you found out about work at your present employer (IF SELF-EMPLOYED:  current organization)?  Please tell me all the ways that you found out about this job.
D. From an acquaintance.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1279/vshow}{GSS Data Explorer page for \code{fndjob4}}.
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{numunemp}},
\code{\link{orgfin}},
\code{\link{totunemp}},
\code{\link{wrkyears}}

Other Employment: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{numunemp}},
\code{\link{totunemp}},
\code{\link{wrkyears}}
}
\concept{Employment}
\concept{Work Organizations}
\keyword{variable}
