% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_137.R
\name{fndjob9}
\alias{fndjob9}
\title{Heard job other (specify)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fndjob9
}
\details{
Question There are many ways people hear about jobs -- such as from ads, employment agencies, or from other people. Could you tell me how you found out about work at your present employer (IF SELF-EMPLOYED:  current organization)?  Please tell me all the ways that you found out about this job.
I. Other (SPECIFY).
}
\section{Values}{

\itemize{
\item \code{1} union
\item \code{2} school
\item \code{3} self-employed,created job
\item \code{4} announcement,newspaper
\item \code{7} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1284/vshow}{GSS Data Explorer page for \code{fndjob9}}.
}

\seealso{
Other Work Organizations: 
\code{\link{drugtest}},
\code{\link{fndjob5}},
\code{\link{fndjob6}},
\code{\link{fndjob7}},
\code{\link{fndjob8}},
\code{\link{intltest}},
\code{\link{ltrsref}},
\code{\link{physical}},
\code{\link{skiltest}}
}
\concept{Work Organizations}
\keyword{variable}
