% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{follow}
\alias{follow}
\title{Importance of following church teachings}
\source{
General Social Survey https://gss.norc.org
}
\description{
follow
}
\details{
Question People have many differing views about what makes a person a good Christian or Jew. Please tell me how important is each of the following to you. Once again, please indicate where you would place your feelings on a scale from 1 to 5. How important is it …
C. To follow faithfully the teachings of their church or synagogue.
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{5} not very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1091/vshow}{GSS Data Explorer page for \code{follow}}.
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}

Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{goownway}},
\code{\link{madatgod}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
