% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{forcult}
\alias{forcult}
\title{Exposure to foreign film damage to our culture}
\source{
General Social Survey https://gss.norc.org
}
\description{
forcult
}
\details{
Question How much do you agree or disagree with the following statements? (Please, select one response)
F. Increased exposure to foreign films, music, and books is damaging our national and local cultures.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4888/vshow}{GSS Data Explorer page for \code{forcult}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{amproud1}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta2a}},
\code{\link{nafta3}}
}
\concept{ISSP National Identity}
\keyword{variable}
