% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{forgive1}
\alias{forgive1}
\title{Have forgiven self for mistakes}
\source{
General Social Survey https://gss.norc.org
}
\description{
forgive1
}
\details{
Question Because of your religious or spiritual beliefs have you always, almost always, often, seldom, or never felt the following ways:
A. I have forgiven myself for things that I have done wrong.
}
\section{Values}{

\itemize{
\item \verb{[1]} alwys/almst alwy
\item \verb{[2]} often
\item \verb{[3]} seldom
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2105/vshow}{GSS Data Explorer page for \code{forgive1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab alwys/almst alwy \tab often \tab seldom \tab never \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 625 \tab 512 \tab 191 \tab 84 \tab 4 \tab 29 \tab 2832 \cr
   Total \tab 1387 \tab 625 \tab 512 \tab 191 \tab 84 \tab 4 \tab 29 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{relcmfrt}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
