% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{forgive2}
\alias{forgive2}
\title{Have forgiven those who hurt me}
\source{
General Social Survey https://gss.norc.org
}
\description{
forgive2
}
\details{
Question Because of your religious or spiritual beliefs have you always, almost always, often, seldom, or never felt the following ways:
B. I have forgiven those who hurt me.
}
\section{Values}{

\itemize{
\item \code{1} alwys/almst alwy
\item \code{2} often
\item \code{3} seldom
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2106/vshow}{GSS Data Explorer page for \code{forgive2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab alwys/almst alwy \tab often \tab seldom \tab never \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 669 \tab 547 \tab 141 \tab 60 \tab 3 \tab 25 \tab 2832 \cr
   Total \tab 1387 \tab 669 \tab 547 \tab 141 \tab 60 \tab 3 \tab 25 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive1}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{relcmfrt}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
