% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{forgive3}
\alias{forgive3}
\title{Know that god forgives me}
\source{
General Social Survey https://gss.norc.org
}
\description{
forgive3
}
\details{
Question Because of your religious or spiritual beliefs have you always, almost always, often, seldom, or never felt the following ways:
C. I know thatd forgives me.
}
\section{Values}{

\itemize{
\item \verb{[1]} alwys/almst alwy
\item \verb{[2]} often
\item \verb{[3]} seldom
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2107/vshow}{GSS Data Explorer page for \code{forgive3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab alwys/almst alwy \tab often \tab seldom \tab never \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 1044 \tab 245 \tab 46 \tab 68 \tab 5 \tab 37 \tab 2832 \cr
   Total \tab 1387 \tab 1044 \tab 245 \tab 46 \tab 68 \tab 5 \tab 37 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{relcmfrt}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
