% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{form}
\alias{form}
\title{Form of split questionnaire asked}
\source{
General Social Survey https://gss.norc.org
}
\description{
form
}
\details{
Question A split ballot was used in the 1973, 1974, 1976, 1978, 1980, and 1982+. The variant or "Y" wordings appear on Form 2; the "Z" wordings appear on Form 3; and the standard GSS wordings appear on Form 1. See Appendix P for what variables were on what form. See   RELIGSP-FEELREL for use of weight variable. See Appendix T, GSS Methodological Report No. 36. See also  MARELKID and PARELKID.
}
\section{Values}{

\itemize{
\item \code{1} standard (x)
\item \code{2} alternate (y)
\item \code{3} alternate (z)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/697/vshow}{GSS Data Explorer page for \code{form}}.
}

\seealso{
Other Core: 
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abspno}},
\code{\link{comprend}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{phone}},
\code{\link{radiohrs}}
}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
