% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{formldys}
\alias{formldys}
\title{The number of days in formal training}
\source{
General Social Survey https://gss.norc.org
}
\description{
formldys
}
\details{
Question D. In total, did the total number of days add up to...
}
\section{Values}{

\itemize{
\item \code{1} a month or less
\item \code{2} 1-3 months
\item \code{3} 3-6 months
\item \code{4} 6 months-1 yr
\item \code{5} over a year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2152/vshow}{GSS Data Explorer page for \code{formldys}}.
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
