% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{formlhrs}
\alias{formlhrs}
\title{The number of hrs in formal training}
\source{
General Social Survey https://gss.norc.org
}
\description{
formlhrs
}
\details{
Question E. In total, did the total number of hours add up to...
}
\section{Values}{

\itemize{
\item \code{1} 35 hours or less
\item \code{2} 35 hours-70 hrs
\item \code{3} over 70 hours
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2153/vshow}{GSS Data Explorer page for \code{formlhrs}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 35 hours or less \tab 35 hours-70 hrs \tab over 70 hours \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2377 \tab 85 \tab 68 \tab 241 \tab 2 \tab 59 \tab 2832 \cr
   Total \tab 2377 \tab 85 \tab 68 \tab 241 \tab 2 \tab 59 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
