% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{forspy}
\alias{forspy}
\title{Foreign spies more threat 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
forspy
}
\details{
Question Are the following threats to the United States greater, about the same, or less today than they were 10 years ago?
B. Spying by foreign agents
}
\section{Values}{

\itemize{
\item \code{1} greater
\item \code{2} about the same
\item \code{3} less
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5258/vshow}{GSS Data Explorer page for \code{forspy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab greater \tab about the same \tab less \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 496 \tab 584 \tab 167 \tab 156 \tab 16 \tab 2817 \cr
   Total \tab 1398 \tab 496 \tab 584 \tab 167 \tab 156 \tab 16 \tab 2817 \cr
}
}

\seealso{
Other National Security: 
\code{\link{camwrk}},
\code{\link{forsteal}},
\code{\link{frterror}},
\code{\link{nuclrwar}},
\code{\link{srchwrk}},
\code{\link{taphme}},
\code{\link{tapwrk}},
\code{\link{usspy}},
\code{\link{usterror}}
}
\concept{National Security}
\keyword{variable}
