% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{frdenjoy}
\alias{frdenjoy}
\title{Should close friend be someone who is enjoyable}
\source{
General Social Survey https://gss.norc.org
}
\description{
frdenjoy
}
\details{
Question People look for various things in a close friend and can differ on how important or not some things are for them. please select a response to say how important or not it is for close friends of yours to be each of the following:
D. Someone who is enjoyable company
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely important
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not too important
\item \verb{[5]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4228/vshow}{GSS Data Explorer page for \code{frdenjoy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab extremely important \tab very important \tab fairly important \tab not too important \tab not at all important \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 548 \tab 467 \tab 109 \tab 12 \tab 5 \tab 8 \tab 2765 \cr
   Total \tab 1616 \tab 548 \tab 467 \tab 109 \tab 12 \tab 5 \tab 8 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{firstyou}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{helpjob}},
\code{\link{kidpars}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{talkedto}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
