% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{frdhelps}
\alias{frdhelps}
\title{Should close friend be someone who helps r}
\source{
General Social Survey https://gss.norc.org
}
\description{
frdhelps
}
\details{
Question People look for various things in a close friend and can differ on how important or not some things are for them. please select a response to say how important or not it is for close friends of yours to be each of the following:
B. Someone who helps me get things done
}
\section{Values}{

\itemize{
\item \code{1} extremely important
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not too important
\item \code{5} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4226/vshow}{GSS Data Explorer page for \code{frdhelps}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{helpjob}},
\code{\link{kidpars}},
\code{\link{learnjob}},
\code{\link{lentto}},
\code{\link{talkedto}}

Other Friends: 
\code{\link{frdenjoy}},
\code{\link{frdknows}},
\code{\link{frdthink}}
}
\concept{Friends}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
