% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{freemns3}
\alias{freemns3}
\title{What does freedom mean to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
freemns3
}
\details{
Question The next questions are about freedom in American today. Freedom means many different things to people. When you think about freedom, what comes to mind? Can you tell me in a couple of sentences what does freedom mean to you?
C. Third distinct subject mentioned
}
\section{Values}{

\itemize{
\item \code{1} want:doing what i want
\item \code{2} autonomy:self-realization
\item \code{3} choices:making decision
\item \code{4} rights:religion and speech
\item \code{5} movement:movement and travel
\item \code{6} negative:slave surbordinate discrimination
\item \code{7} citizen:political participation
\item \code{8} security:economic independence
\item \code{9} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2659/vshow}{GSS Data Explorer page for \code{freemns3}}.
}

\seealso{
Other Freedom: 
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}}
}
\concept{Freedom}
\keyword{variable}
