% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{freeprss}
\alias{freeprss}
\title{Press should be completely free}
\source{
General Social Survey https://gss.norc.org
}
\description{
freeprss
}
\details{
Question Many people think it is important to have complete freedom of the press and news media in the United States. Others think that the press too often invades the privacy of public figures–like Senators or members of Congress by printing stories that contain personal details about their private lives. Which of the three statements on this Card comes closest to your feelings about balancing freedom of the press and the right to privacy?
READ STATEMENTS 1-3 ONLY. CIRCLE ONLY ONE CODE.
}
\section{Values}{

\itemize{
\item \code{1} there should be complete freedom of the press, even if
\item \code{2} the press should develop a code of ethics to keep it
\item \code{3} the government should keep the press from printing
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2674/vshow}{GSS Data Explorer page for \code{freeprss}}.
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{freenow}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}},
\code{\link{wlthpov}}
}
\concept{Freedom}
\keyword{variable}
