% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{fricall}
\alias{fricall}
\title{Contact with closest friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
fricall
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to…
F. And how often do you have any other contact with this friend besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \verb{[1]} daily
\item \verb{[2]} several x week
\item \verb{[3]} once a week
\item \verb{[4]} once a month
\item \verb{[5]} several x year
\item \verb{[6]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4168/vshow}{GSS Data Explorer page for \code{fricall}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab no answer \tab Total \cr
   1986 \tab 92 \tab 285 \tab 312 \tab 277 \tab 247 \tab 104 \tab 141 \tab 12 \tab 1470 \cr
   Total \tab 92 \tab 285 \tab 312 \tab 277 \tab 247 \tab 104 \tab 141 \tab 12 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
