% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{frineigh}
\alias{frineigh}
\title{Number of close friends are neighbors}
\source{
General Social Survey https://gss.norc.org
}
\description{
frineigh
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to… How many close friends would you say you have?
B. How many of these close friends are your neighbors now?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4164/vshow}{GSS Data Explorer page for \code{frineigh}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}},
\code{\link{relcall}}

Other Friends: 
\code{\link{fricall}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}}

Other Social Networks: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}}
}
\concept{Friends}
\concept{ISSP Social Networks & Support Systems}
\concept{Social Networks}
\keyword{variable}
