% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{frivisit}
\alias{frivisit}
\title{How often visit closest friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
frivisit
}
\details{
Question Thinking now of close friends - not your husband or wife or partner or family members - but people you feel fairly close to…
D. How often do you see or visit with your friend (the friend you feel closest to)?
}
\section{Values}{

\itemize{
\item \code{1} lives in hshld
\item \code{2} daily
\item \code{3} several x week
\item \code{4} once a week
\item \code{5} once a month
\item \code{6} several x year
\item \code{7} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4166/vshow}{GSS Data Explorer page for \code{frivisit}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab lives in hshld \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab don't know \tab no answer \tab Total \cr
   1986 \tab 77 \tab 14 \tab 258 \tab 317 \tab 316 \tab 232 \tab 160 \tab 86 \tab 1 \tab 9 \tab 1470 \cr
   Total \tab 77 \tab 14 \tab 258 \tab 317 \tab 316 \tab 232 \tab 160 \tab 86 \tab 1 \tab 9 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{friwork}},
\code{\link{relcall}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
