% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{frndawk}
\alias{frndawk}
\title{Discussing price is awkward among friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndawk
}
\details{
Question When friends buy or sell something to friends, discussing price is usually awkward or puts somebody on the spot.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1864/vshow}{GSS Data Explorer page for \code{frndawk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 269 \tab 766 \tab 165 \tab 191 \tab 25 \tab 22 \tab 6 \tab 2904 \cr
   Total \tab 1460 \tab 269 \tab 766 \tab 165 \tab 191 \tab 25 \tab 22 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
