% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_119.R
\name{frndcon2}
\alias{frndcon2}
\title{Does friend 2 belong to r's congregation}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndcon2
}
\details{
Question Many people have some good friends they feel close to. Who are your good friends (other than your spouse)?  Just tell me their first names.
(LIST NAMES IN GRID BELOW). Is there anyone else?
NAME 1

Now let's go back and talk about (NAME). (REPEAT FOR EACH FRIEND)
Is (NAME) a member of your congregation?
2. Second Person Member of Respondent's Congregation
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} r isnt in congregtn
\item \code{4} friend, no info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1104/vshow}{GSS Data Explorer page for \code{frndcon2}}.
}

\seealso{
Other Religion: 
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{frndcon1}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}}
}
\concept{Religion}
\keyword{variable}
