% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_120.R
\name{frndden3}
\alias{frndden3}
\title{What is friend 3s specific denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndden3
}
\details{
Question What specific denomination is that, if any?
(PROBE FOR EXACT DENOMINATION. REFER TO PROTESTANT DENOMINATION LIST AND CIRCLE APPROPRIATE CODE BELOW).
3. Denomination of Third Person
}
\section{Values}{

\itemize{
\item \verb{[10]} am baptist asso
\item \verb{[11]} am bapt ch in usa
\item \verb{[12]} nat bapt conv of am
\item \verb{[13]} nat bapt conv usa
\item \verb{[14]} southern baptist
\item \verb{[15]} other baptists
\item \verb{[18]} baptist-dk which
\item \verb{[20]} afr meth episcopal
\item \verb{[21]} afr meth ep zion
\item \verb{[22]} united methodist
\item \verb{[23]} other methodist
\item \verb{[28]} methodist-dk which
\item \verb{[30]} am lutheran
\item \verb{[31]} luth ch in america
\item \verb{[32]} lutheran-mo synod
\item \verb{[33]} wi evan luth synod
\item \verb{[34]} other lutheran
\item \verb{[35]} evangelical lutheran
\item \verb{[38]} lutheran-dk which
\item \verb{[40]} presbyterian c in us
\item \verb{[41]} united pres ch in usa
\item \verb{[42]} other presbyterian
\item \verb{[43]} presbyt ch usa
\item \verb{[48]} presbyterian-dk wh
\item \verb{[50]} episcopal
\item \verb{[60]} other
\item \verb{[70]} no denomination
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1115/vshow}{GSS Data Explorer page for \code{frndden3}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab am baptist asso \tab am bapt ch in usa \tab nat bapt conv of am \tab nat bapt conv usa \tab southern baptist \tab other baptists \tab baptist-dk which \tab afr meth episcopal \tab afr meth ep zion \tab united methodist \tab other methodist \tab methodist-dk which \tab am lutheran \tab luth ch in america \tab lutheran-mo synod \tab wi evan luth synod \tab other lutheran \tab lutheran-dk which \tab presbyterian c in us \tab united pres ch in usa \tab other presbyterian \tab presbyterian-dk wh \tab episcopal \tab other \tab no denomination \tab don't know \tab no answer \tab evangelical lutheran \tab Total \cr
   1988 \tab 908 \tab 11 \tab 5 \tab 1 \tab 3 \tab 83 \tab 21 \tab 53 \tab 3 \tab 1 \tab 43 \tab 5 \tab 22 \tab 10 \tab 5 \tab 12 \tab 1 \tab 1 \tab 17 \tab 3 \tab 9 \tab 1 \tab 15 \tab 10 \tab 90 \tab 50 \tab 10 \tab 88 \tab - \tab 1481 \cr
   1998 \tab 2481 \tab 8 \tab 3 \tab 1 \tab 1 \tab 32 \tab 5 \tab 25 \tab - \tab - \tab 17 \tab 4 \tab 10 \tab 6 \tab - \tab 7 \tab - \tab 4 \tab 6 \tab 1 \tab 2 \tab 2 \tab 8 \tab 12 \tab 38 \tab 45 \tab 4 \tab 108 \tab 2 \tab 2832 \cr
   Total \tab 3389 \tab 19 \tab 8 \tab 2 \tab 4 \tab 115 \tab 26 \tab 78 \tab 3 \tab 1 \tab 60 \tab 9 \tab 32 \tab 16 \tab 5 \tab 19 \tab 1 \tab 5 \tab 23 \tab 4 \tab 11 \tab 3 \tab 23 \tab 22 \tab 128 \tab 95 \tab 14 \tab 196 \tab 2 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}}
}
\concept{Religion}
\keyword{variable}
