% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{frndking}
\alias{frndking}
\title{Contrast between friend and king}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndking
}
\details{
Question There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? (HAND CARD AE) The first set of contrasting images shows Mother at 1 on the scale and Father at 7. If you imagine God as a Mother you would place yourself at 1. If you imagine God as a Father, you would place yourself at 7. If you imagine God as somewhere between Mother and Father, you would place yourself at 2, 3, 4, 5, or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F, SUBSTITUTING IMAGES A-F FOR "MOTHER" AND "FATHER".) Where would you place your image of God on the scale for . . . READ EACH SET OF IMAGES AND CIRCLE ONE CODE FOR EACH.
}
\section{Values}{

\itemize{
\item \code{1} friend
\item \code{7} king
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/337/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab friend \tab king \tab no answer \tab don't know \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 77 \tab 41 \tab 258 \tab 53 \tab 29 \tab 293 \tab 180 \tab 58 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 114 \tab 82 \tab 406 \tab 82 \tab 70 \tab 433 \tab 280 \tab 50 \tab 17 \tab - \tab 1534 \cr
   1986 \tab - \tab 105 \tab 77 \tab 383 \tab 53 \tab 69 \tab 469 \tab 257 \tab 47 \tab 10 \tab - \tab 1470 \cr
   1987 \tab - \tab 86 \tab 71 \tab 497 \tab 72 \tab 67 \tab 552 \tab 385 \tab 73 \tab 16 \tab - \tab 1819 \cr
   1988 \tab - \tab 95 \tab 64 \tab 420 \tab 69 \tab 42 \tab 419 \tab 290 \tab 54 \tab 28 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 72 \tab 40 \tab 286 \tab 54 \tab 37 \tab 268 \tab 207 \tab 39 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 67 \tab 41 \tab 258 \tab 38 \tab 29 \tab 263 \tab 180 \tab 47 \tab 5 \tab - \tab 1372 \cr
   1991 \tab 158 \tab 112 \tab 77 \tab 341 \tab 81 \tab 71 \tab 277 \tab 193 \tab 207 \tab - \tab - \tab 1517 \cr
   1993 \tab 526 \tab 79 \tab 59 \tab 309 \tab 44 \tab 45 \tab 278 \tab 212 \tab 51 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 110 \tab 81 \tab 585 \tab 94 \tab 76 \tab 525 \tab 415 \tab 69 \tab 22 \tab - \tab 2992 \cr
   1996 \tab 1913 \tab 50 \tab 48 \tab 304 \tab 66 \tab 41 \tab 269 \tab 171 \tab 36 \tab 6 \tab - \tab 2904 \cr
   1998 \tab 1387 \tab 75 \tab 70 \tab 430 \tab 87 \tab 79 \tab 322 \tab 317 \tab 50 \tab 15 \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 69 \tab 78 \tab 382 \tab 86 \tab 71 \tab 253 \tab 311 \tab 31 \tab 84 \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 51402 \tab 1111 \tab 829 \tab 4859 \tab 879 \tab 726 \tab 4621 \tab 3398 \tab 812 \tab 209 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{crtrheal}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}},
\code{\link{rdeemlib}}

Other Religion: 
\code{\link{crtrheal}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}},
\code{\link{rdeemlib}}

Other Split Ballots: 
\code{\link{crtrheal}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{rdeemlib}}
}
\concept{Core}
\concept{Religion}
\concept{Split Ballots}
\keyword{variable}
