% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{frndknow}
\alias{frndknow}
\title{How many of r's friends know each other?}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndknow
}
\details{
Question Some people have friends who mostly know one another. Other people have friends who don't know one another. would you say that all of your friends know one another, most of your friends know one another, only a few of your friends know one another, or none of your friends know one another?
}
\section{Values}{

\itemize{
\item \verb{[1]} all know each other
\item \verb{[2]} most know each other
\item \verb{[3]} few know each other
\item \verb{[4]} none know each other
\item \verb{[5]} no friends(vol.)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/962/vshow}{GSS Data Explorer page for \code{frndknow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all know each other \tab most know each other \tab few know each other \tab none know each other \tab no friends(vol.) \tab don't know \tab no answer \tab Total \cr
   1985 \tab - \tab 183 \tab 742 \tab 536 \tab 52 \tab 4 \tab 15 \tab 2 \tab 1534 \cr
   Total \tab 0 \tab 183 \tab 742 \tab 536 \tab 52 \tab 4 \tab 15 \tab 2 \tab 1534 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{talkpol}},
\code{\link{talkpol3}}
}
\concept{Social Networks}
\keyword{variable}
