% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_120.R
\name{frndoth1}
\alias{frndoth1}
\title{What is friend 1s other prot denom}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndoth1
}
\details{
Question What specific denomination is that, if any?
\enumerate{
\item Denomination of first friend
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1118/vshow}{GSS Data Explorer page for \code{frndoth1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 100 \tab 103 \tab 116 \tab 12 \tab 128 \tab 129 \tab 18 \tab 23 \tab 27 \tab 3 \tab 32 \tab 33 \tab 35 \tab 36 \tab 37 \tab 40 \tab 43 \tab 44 \tab 47 \tab 51 \tab 52 \tab 54 \tab 56 \tab 58 \tab 59 \tab 60 \tab 61 \tab 64 \tab 65 \tab 67 \tab 68 \tab 71 \tab 77 \tab 78 \tab 81 \tab 83 \tab 9 \tab 93 \tab 95 \tab 97 \tab don't know \tab no answer \tab 110 \tab 24 \tab 31 \tab 42 \tab 69 \tab 80 \tab Total \cr
   1988 \tab 1268 \tab 1 \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 8 \tab 5 \tab 1 \tab 7 \tab 1 \tab 4 \tab 1 \tab 6 \tab 3 \tab 1 \tab 4 \tab 5 \tab 10 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 9 \tab 3 \tab 3 \tab 3 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 97 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1998 \tab 2677 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 106 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2832 \cr
   Total \tab 3945 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 11 \tab 9 \tab 1 \tab 8 \tab 1 \tab 6 \tab 2 \tab 6 \tab 3 \tab 1 \tab 6 \tab 9 \tab 10 \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 11 \tab 3 \tab 5 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 203 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndoth2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}}
}
\concept{Religion}
\keyword{variable}
