% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_121.R
\name{frndoth3}
\alias{frndoth3}
\title{What is friend 3s other prot denom}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndoth3
}
\details{
Question What specific denomination is that, if any?
3. Denomination of third friend
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1120/vshow}{GSS Data Explorer page for \code{frndoth3}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 12 \tab 14 \tab 18 \tab 20 \tab 23 \tab 33 \tab 35 \tab 36 \tab 40 \tab 44 \tab 51 \tab 52 \tab 54 \tab 56 \tab 58 \tab 59 \tab 61 \tab 64 \tab 65 \tab 67 \tab 68 \tab 71 \tab 74 \tab 77 \tab 78 \tab 80 \tab 81 \tab 9 \tab 93 \tab 95 \tab 97 \tab don't know \tab no answer \tab 103 \tab 138 \tab 149 \tab 150 \tab 31 \tab 63 \tab 69 \tab 99 \tab Total \cr
   1988 \tab 1303 \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 4 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 8 \tab 10 \tab 1 \tab 3 \tab 4 \tab 1 \tab 10 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 91 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1998 \tab 2686 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab 110 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2832 \cr
   Total \tab 3989 \tab 12 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 8 \tab 4 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 9 \tab 10 \tab 1 \tab 3 \tab 4 \tab 1 \tab 13 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 201 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}}
}
\concept{Religion}
\keyword{variable}
