% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_121.R
\name{frndoth5}
\alias{frndoth5}
\title{What is friend 5s other prot denom}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndoth5
}
\details{
Question What specific denomination is that, if any?
5. Denomination of fifth friend
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1122/vshow}{GSS Data Explorer page for \code{frndoth5}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 103 \tab 147 \tab 20 \tab 31 \tab 33 \tab 35 \tab 40 \tab 56 \tab 64 \tab 69 \tab 80 \tab 97 \tab no answer \tab Total \cr
   1998 \tab 2735 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 83 \tab 2832 \cr
   Total \tab 2735 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 83 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndrac1}},
\code{\link{frndrac2}}
}
\concept{Religion}
\keyword{variable}
