% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_119.R
\name{frndrel2}
\alias{frndrel2}
\title{What is friend 2s religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndrel2
}
\details{
Question What''s is (NAME''S) religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?
2. Religion of Second Person
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1109/vshow}{GSS Data Explorer page for \code{frndrel2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1988 \tab 160 \tab 579 \tab 297 \tab 34 \tab 102 \tab 46 \tab 183 \tab 80 \tab 1481 \cr
   1998 \tab 1877 \tab 308 \tab 219 \tab 18 \tab 92 \tab 28 \tab 204 \tab 86 \tab 2832 \cr
   Total \tab 2037 \tab 887 \tab 516 \tab 52 \tab 194 \tab 74 \tab 387 \tab 166 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndrel1}}
}
\concept{Religion}
\keyword{variable}
