% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_366.R
\name{frstmet3}
\alias{frstmet3}
\title{R first met the named alter in the group}
\source{
General Social Survey https://gss.norc.org
}
\description{
frstmet3
}
\details{
Question Did you first meet (NAME) in one of the groups in the previous question, or did you first meet somewhere else?
C. Did you meet the third name in one of these groups or did you meet somewhere else?
}
\section{Values}{

\itemize{
\item \code{1} yes, one of these groups
\item \code{2} no, somewhere else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3305/vshow}{GSS Data Explorer page for \code{frstmet3}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet2}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{relneg}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}

Other Social Networks: 
\code{\link{frstmet2}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Social Networks}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
