% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_156.R
\name{fstspedc}
\alias{fstspedc}
\title{Educ of first spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
fstspedc
}
\details{
Question RESPONDENT'S FIRST SPOUSE'S EDUCATION
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1477/vshow}{GSS Data Explorer page for \code{fstspedc}}.
}

\seealso{
Other Family Mobility: 
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{fstspdeg}},
\code{\link{fstspyr}},
\code{\link{marnum}}

Other Education: 
\code{\link{fstspdeg}}
}
\concept{Education}
\concept{Family Mobility}
\keyword{variable}
