% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_156.R
\name{fstspedc}
\alias{fstspedc}
\title{Educ of first spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
fstspedc
}
\details{
Question RESPONDENT'S FIRST SPOUSE'S EDUCATION
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1477/vshow}{GSS Data Explorer page for \code{fstspedc}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 3 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2432 \tab 30 \tab 23 \tab 227 \tab 24 \tab 49 \tab 12 \tab 47 \tab 6 \tab 8 \tab 4 \tab 11 \tab 1 \tab 4 \tab 10 \tab 9 \tab 26 \tab 15 \tab 50 \tab 4 \tab 2992 \cr
   Total \tab 2432 \tab 30 \tab 23 \tab 227 \tab 24 \tab 49 \tab 12 \tab 47 \tab 6 \tab 8 \tab 4 \tab 11 \tab 1 \tab 4 \tab 10 \tab 9 \tab 26 \tab 15 \tab 50 \tab 4 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{fstspdeg}},
\code{\link{fstspyr}},
\code{\link{marnum}}
}
\concept{Family Mobility}
\keyword{variable}
