% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_673.R
\name{ftruse}
\alias{ftruse}
\title{Likelihood r uses internet in next 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
ftruse
}
\details{
Question How likely is it that you will personally use the Internet in the next 12 months?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} rather likely
\item \code{3} rather unlikely
\item \code{4} very unlikely
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8144/vshow}{GSS Data Explorer page for \code{ftruse}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intaccprxy}},
\code{\link{inthome}},
\code{\link{intprfssnl}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{prxyuse}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
