% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{fucitznnv}
\alias{fucitznnv}
\title{Is r planning/appling for us citizenship or not (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
fucitznnv
}
\details{
Question Are you…
}
\section{Values}{

\itemize{
\item \code{1} currently applying for u.s. citizenship
\item \code{2} planning to apply for u.s. citizenship
\item \code{3} not planning to apply to u.s. citizenship
\item \code{4} not eligible to become a u.s. citizen (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7621/vshow}{GSS Data Explorer page for \code{fucitznnv}}.
}

\keyword{variable}
