% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{fucitznv}
\alias{fucitznv}
\title{Is r planning/appling for us citizenship or not (with volunteered response on we}
\source{
General Social Survey https://gss.norc.org
}
\description{
fucitznv
}
\details{
Question Are you…
}
\section{Values}{

\itemize{
\item \code{1} currently applying for u.s. citizenship
\item \code{2} planning to apply for u.s. citizenship
\item \code{3} not planning to apply to u.s. citizenship
\item \code{4} not eligible to become a u.s. citizen (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7601/vshow}{GSS Data Explorer page for \code{fucitznv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab currently applying for u.s. citizenship \tab planning to apply for u.s. citizenship \tab not planning to apply to u.s. citizenship \tab not eligible to become a u.s. citizen (phone mode only: volunteered) \tab skipped on web \tab Total \cr
   2021 \tab 3939 \tab 8 \tab 50 \tab 21 \tab 10 \tab 4 \tab 4032 \cr
   2022 \tab 3507 \tab 4 \tab 18 \tab 6 \tab 7 \tab 2 \tab 3544 \cr
   Total \tab 7446 \tab 12 \tab 68 \tab 27 \tab 17 \tab 6 \tab 7576 \cr
}
}

\keyword{variable}
