% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{futenrgy}
\alias{futenrgy}
\title{U.s. energy priorities}
\source{
General Social Survey https://gss.norc.org
}
\description{
futenrgy
}
\details{
Question G. To which of the following should America give priority in order to meet its future energy needs?
}
\section{Values}{

\itemize{
\item \code{1} coal, oil and natural gas
\item \code{2} nuclear power
\item \code{3} solar, wind or water power
\item \code{4} fuels made from crops
\item \code{5} fuels made from waste
\item \code{6} none of them
\item \code{7} can't choose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4807/vshow}{GSS Data Explorer page for \code{futenrgy}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
