% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{futenrgy}
\alias{futenrgy}
\title{U.s. energy priorities}
\source{
General Social Survey https://gss.norc.org
}
\description{
futenrgy
}
\details{
Question G. To which of the following should America give priority in order to meet its future energy needs?
}
\section{Values}{

\itemize{
\item \verb{[1]} coal, oil and natural gas
\item \verb{[2]} nuclear power
\item \verb{[3]} solar, wind or water power
\item \verb{[4]} fuels made from crops
\item \verb{[5]} fuels made from waste
\item \verb{[6]} none of them
\item \verb{[7]} can't choose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4807/vshow}{GSS Data Explorer page for \code{futenrgy}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab coal, oil and natural gas \tab nuclear power \tab solar, wind or water power \tab fuels made from crops \tab fuels made from waste \tab none of them \tab can't choose \tab don't know \tab no answer \tab Total \cr
   2010 \tab 614 \tab 170 \tab 144 \tab 650 \tab 57 \tab 162 \tab 25 \tab 204 \tab 12 \tab 6 \tab 2044 \cr
   Total \tab 614 \tab 170 \tab 144 \tab 650 \tab 57 \tab 162 \tab 25 \tab 204 \tab 12 \tab 6 \tab 2044 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
