% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{futpromo}
\alias{futpromo}
\title{How likely you will be promoted next five years}
\source{
General Social Survey https://gss.norc.org
}
\description{
futpromo
}
\details{
Question In the next five years, how likely are you to be promoted?  Would you say very likely, likely, not very likely, or not likely at all?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} likely
\item \code{3} not very likely
\item \code{4} not likely at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1303/vshow}{GSS Data Explorer page for \code{futpromo}}.
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}

Other Employment: 
\code{\link{advances}},
\code{\link{imppromo}},
\code{\link{numpromo}},
\code{\link{promoted}}
}
\concept{Employment}
\concept{Work Organizations}
\keyword{variable}
