% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{futrup}
\alias{futrup}
\title{Degree economy will improve in next year}
\source{
General Social Survey https://gss.norc.org
}
\description{
futrup
}
\details{
Question What about the next 12 months? Do you expect the national economy to get better, get worse, or stay about the same?
A. Would you say much better or somewhat better?
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} somewhat better
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1645/vshow}{GSS Data Explorer page for \code{futrup}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab much better \tab somewhat better \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2566 \tab 60 \tab 351 \tab 2 \tab 13 \tab 2992 \cr
   Total \tab 2566 \tab 60 \tab 351 \tab 2 \tab 13 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}}
}
\concept{Multiculturalism}
\keyword{variable}
