% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{gavepol}
\alias{gavepol}
\title{Has r given money to candidate or cause}
\source{
General Social Survey https://gss.norc.org
}
\description{
gavepol
}
\details{
Question In the past three or four years, have you contributed money to a political party or candidate or to any other political cause?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1040/vshow}{GSS Data Explorer page for \code{gavepol}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab - \tab 390 \tab 1417 \tab 12 \tab 1819 \cr
   Total \tab 0 \tab 390 \tab 1417 \tab 12 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{blkgains}},
\code{\link{blkinflu}},
\code{\link{fegains}},
\code{\link{feinflu}}
}
\concept{Socio-political participation}
\keyword{variable}
