% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{gdjobsec}
\alias{gdjobsec}
\title{Job security is good}
\source{
General Social Survey https://gss.norc.org
}
\description{
gdjobsec
}
\details{
Question L. At work, job security is good.
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3009/vshow}{GSS Data Explorer page for \code{gdjobsec}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 971 \tab 597 \tab 165 \tab 76 \tab 15 \tab 43 \tab 2812 \cr
   Total \tab 945 \tab 971 \tab 597 \tab 165 \tab 76 \tab 15 \tab 43 \tab 2812 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
