% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{gdjobsec}
\alias{gdjobsec}
\title{Job security is good}
\source{
General Social Survey https://gss.norc.org
}
\description{
gdjobsec
}
\details{
Question L. At work, job security is good.
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3009/vshow}{GSS Data Explorer page for \code{gdjobsec}}.
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}

Other Job: 
\code{\link{actupset}},
\code{\link{hotargus}},
\code{\link{lookaway}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Job}
\concept{Workplace Conflict}
\keyword{variable}
