% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_583.R
\name{gender3}
\alias{gender3}
\title{Gender of 3rd person}
\source{
General Social Survey https://gss.norc.org
}
\description{
gender3
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
2. Gender of third person
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[3]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5345/vshow}{GSS Data Explorer page for \code{gender3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab male \tab female \tab no answer \tab don't know \tab not available in this release \tab 3 \tab skipped on web \tab Total \cr
   1975 \tab 656 \tab 463 \tab 371 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1134 \tab 200 \tab 164 \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 724 \tab 436 \tab 368 \tab 2 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 749 \tab 419 \tab 361 \tab 3 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 780 \tab 362 \tab 322 \tab 4 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 991 \tab 416 \tab 445 \tab 8 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 797 \tab 394 \tab 400 \tab 8 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 783 \tab 345 \tab 341 \tab 4 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 844 \tab 352 \tab 336 \tab 2 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 768 \tab 354 \tab 347 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 948 \tab 447 \tab 421 \tab 3 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 847 \tab 320 \tab 313 \tab 1 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 834 \tab 386 \tab 305 \tab 12 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 810 \tab 293 \tab 263 \tab 6 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 849 \tab 348 \tab 317 \tab 3 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 895 \tab 364 \tab 344 \tab 3 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1751 \tab 656 \tab 569 \tab 16 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1729 \tab 574 \tab 566 \tab 35 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1713 \tab 557 \tab 546 \tab 16 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1668 \tab 582 \tab 530 \tab 37 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1815 \tab 488 \tab 462 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1780 \tab 553 \tab 472 \tab 7 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2778 \tab 901 \tab 816 \tab 2 \tab 13 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1221 \tab 401 \tab 391 \tab 3 \tab 7 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1319 \tab 364 \tab 352 \tab 4 \tab 5 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1259 \tab 376 \tab 328 \tab 8 \tab 3 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1642 \tab 481 \tab 403 \tab 6 \tab 6 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1919 \tab 513 \tab 417 \tab 10 \tab 8 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1561 \tab 405 \tab 367 \tab 10 \tab 5 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 2841 \tab 355 \tab 335 \tab 4 \tab - \tab - \tab 7 \tab 2 \tab 3544 \cr
   Total \tab 38405 \tab 13105 \tab 11972 \tab 219 \tab 47 \tab 4032 \tab 7 \tab 2 \tab 67789 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{where2}},
\code{\link{where3}}
}
\concept{Household Members and Structure}
\keyword{variable}
