% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_585.R
\name{gender7}
\alias{gender7}
\title{Gender of 7th person}
\source{
General Social Survey https://gss.norc.org
}
\description{
gender7
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
2. Gender of seventh person
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5369/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab female \tab male \tab no answer \tab don't know \tab other \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1437 \tab 29 \tab 24 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1479 \tab 9 \tab 11 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1484 \tab 24 \tab 21 \tab 1 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1492 \tab 15 \tab 23 \tab 2 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1431 \tab 11 \tab 25 \tab 1 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1817 \tab 16 \tab 26 \tab 1 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1559 \tab 18 \tab 21 \tab 1 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1454 \tab 12 \tab 7 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1505 \tab 14 \tab 15 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1441 \tab 12 \tab 17 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1787 \tab 19 \tab 13 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1465 \tab 8 \tab 7 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1498 \tab 21 \tab 16 \tab 2 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1354 \tab 6 \tab 12 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1497 \tab 9 \tab 11 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1597 \tab 4 \tab 5 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2958 \tab 21 \tab 13 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2862 \tab 25 \tab 15 \tab 2 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2801 \tab 17 \tab 14 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2786 \tab 14 \tab 15 \tab 2 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2741 \tab 10 \tab 14 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2786 \tab 16 \tab 10 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4457 \tab 22 \tab 30 \tab - \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 2002 \tab 14 \tab 7 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2003 \tab 15 \tab 23 \tab 2 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 1951 \tab 13 \tab 10 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2515 \tab 11 \tab 11 \tab - \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 2830 \tab 15 \tab 22 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2327 \tab 11 \tab 9 \tab 1 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3981 \tab 22 \tab 28 \tab - \tab - \tab 1 \tab - \tab 4032 \cr
   2022 \tab 3521 \tab 11 \tab 12 \tab - \tab - \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 71419 \tab 464 \tab 487 \tab 16 \tab 3 \tab 1 \tab 3309 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away6}},
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}

Other Family: 
\code{\link{away6}},
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}

Other Household: 
\code{\link{away6}},
\code{\link{away7}},
\code{\link{gender6}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{relate7}},
\code{\link{where6}}
}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
