% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{genderwk}
\alias{genderwk}
\title{Mostly men or women at the workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
genderwk
}
\details{
Question Are the people who work at this location mostly men or women?  Please just tell me the number on the card?
}
\section{Values}{

\itemize{
\item \verb{[1]} all women
\item \verb{[2]} almost all women
\item \verb{[3]} mostly women
\item \verb{[4]} half men/women
\item \verb{[5]} mostly men
\item \verb{[6]} almost all men
\item \verb{[7]} all men
\item \verb{[8]} not sure
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2145/vshow}{GSS Data Explorer page for \code{genderwk}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab all women \tab almost all women \tab mostly women \tab half men/women \tab mostly men \tab almost all men \tab all men \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1455 \tab 75 \tab 124 \tab 234 \tab 434 \tab 268 \tab 118 \tab 73 \tab 20 \tab 31 \tab 2832 \cr
   Total \tab 1455 \tab 75 \tab 124 \tab 234 \tab 434 \tab 268 \tab 118 \tab 73 \tab 20 \tab 31 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}
}
\concept{Jobs}
\keyword{variable}
