% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{genderwk}
\alias{genderwk}
\title{Mostly men or women at the workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
genderwk
}
\details{
Question Are the people who work at this location mostly men or women?  Please just tell me the number on the card?
}
\section{Values}{

\itemize{
\item \code{1} all women
\item \code{2} almost all women
\item \code{3} mostly women
\item \code{4} half men/women
\item \code{5} mostly men
\item \code{6} almost all men
\item \code{7} all men
\item \code{8} not sure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2145/vshow}{GSS Data Explorer page for \code{genderwk}}.
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}

Other Job: 
\code{\link{formltrn}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}
}
\concept{Job}
\concept{Jobs}
\keyword{variable}
