% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{gendtech}
\alias{gendtech}
\title{Are women or men benefiting more from digital tech}
\source{
General Social Survey https://gss.norc.org
}
\description{
gendtech
}
\details{
Question In all countries, there are differences between social groups. Do you think that in America women or men are benefitting more, or both equally, or neither from the use of the Internet and digital technology?
}
\section{Values}{

\itemize{
\item \verb{[1]} women are benefitting more than men
\item \verb{[2]} women and men are benefitting equally
\item \verb{[3]} men are benefitting more than women
\item \verb{[4]} neither men nor women are benefitting
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8110/vshow}{GSS Data Explorer page for \code{gendtech}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab women are benefitting more than men \tab women and men are benefitting equally \tab men are benefitting more than women \tab neither men nor women are benefitting \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 135 \tab 742 \tab 179 \tab 277 \tab 279 \tab 1674 \tab 8 \tab 15 \tab 3309 \cr
   Total \tab 0 \tab 135 \tab 742 \tab 179 \tab 277 \tab 279 \tab 1674 \tab 8 \tab 15 \tab 3309 \cr
}
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{dataprot}},
\code{\link{emonitor1}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{modeorg}},
\code{\link{trcong}},
\code{\link{trppl}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
