% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{geneabrt}
\alias{geneabrt}
\title{Abort defective fetus}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneabrt
}
\details{
Question Suppose a test shows the baby has a serious genetic defect. Would you (yourself want to/ want your partner to) have an abortion if a test shows the baby has a serious genetic defect?
}
\section{Values}{

\itemize{
\item \code{1} have abortion
\item \code{2} no abortion
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/C/- \tab partial \cr
   1996 \tab A/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
   2018 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5089/vshow}{GSS Data Explorer page for \code{geneabrt}}.
}

\seealso{
Other Genes: 
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{geneself}}

Other Genetic Testing: 
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{geneself}}
}
\concept{Abortion}
\concept{Genes}
\concept{Genetic Testing}
\keyword{variable}
