% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{geneabrt2}
\alias{geneabrt2}
\title{R would want to abort fetus if there were genetic defects}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneabrt2
}
\details{
Question Suppose a test shows the baby has a serious genetic defect. Would you (yourself want/want the pregnant woman/want your partner) to have an abortion if a test shows the baby has a serious genetic defect?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7296/vshow}{GSS Data Explorer page for \code{geneabrt2}}.
}

\keyword{variable}
