% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{genecany}
\alias{genecany}
\title{Employers' right to hire by cancertests}
\source{
General Social Survey https://gss.norc.org
}
\description{
genecany
}
\details{
Question B. Suppose workers will be exposed to some cancer-causing materials on the job. In that case, should employers have the right not to hire workers whose tests show they have an inherited tendency to develop certain forms of cancer, or should they not have that right?
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1261/vshow}{GSS Data Explorer page for \code{genecany}}.
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genegets}},
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}

Other Genetic Testing: 
\code{\link{genecanx}},
\code{\link{genegets}}
}
\concept{Genetic Testing}
\concept{Work Organizations}
\keyword{variable}
