% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{genedef2}
\alias{genedef2}
\title{Meaning of serious genetic defect}
\source{
General Social Survey https://gss.norc.org
}
\description{
genedef2
}
\details{
Question When I say "serious genetic defect," what kinds of defects do you think of?  Any others?
RECORD VERBATIM AND CODE FIRST THREE MENTIONS
B. Second Mention
}
\section{Values}{

\itemize{
\item \verb{[1]} downs syndrome
\item \verb{[2]} mental retardation
\item \verb{[3]} spina bifida
\item \verb{[4]} muscular dystrophy
\item \verb{[5]} cystic fibrosis
\item \verb{[6]} cerebral palsy
\item \verb{[7]} crippling diseases
\item \verb{[10]} blind,deaf
\item \verb{[11]} missing limbs
\item \verb{[12]} major deformities
\item \verb{[13]} life-threatening deform
\item \verb{[14]} handicapped
\item \verb{[15]} physical defects
\item \verb{[20]} sickle cell anemia
\item \verb{[21]} rh factor
\item \verb{[22]} hemophilia
\item \verb{[30]} addiction to drugs
\item \verb{[31]} aids
\item \verb{[35]} multiple sclerosis
\item \verb{[60]} heart
\item \verb{[61]} kidneys
\item \verb{[62]} siamese twins
\item \verb{[70]} incurable disease
\item \verb{[71]} tourettes syndrome
\item \verb{[72]} sexual transmitted diseases
\item \verb{[73]} diabetes
\item \verb{[74]} birth defects
\item \verb{[75]} cancer
\item \verb{[76]} psychological problems
\item \verb{[77]} neurological problems
\item \verb{[78]} hydrocephalus
\item \verb{[79]} huntingtons chorea
\item \verb{[80]} tay-sachs
\item \verb{[81]} reyes syndrome
\item \verb{[90]} unable to take care of self
\item \verb{[95]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5091/vshow}{GSS Data Explorer page for \code{genedef2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab downs syndrome \tab mental retardation \tab spina bifida \tab muscular dystrophy \tab cystic fibrosis \tab cerebral palsy \tab crippling diseases \tab blind,deaf \tab missing limbs \tab major deformities \tab life-threatening deform \tab handicapped \tab physical defects \tab sickle cell anemia \tab rh factor \tab hemophilia \tab addiction to drugs \tab aids \tab multiple sclerosis \tab heart \tab siamese twins \tab incurable disease \tab tourettes syndrome \tab sexual transmitted diseases \tab diabetes \tab birth defects \tab cancer \tab psychological problems \tab neurological problems \tab hydrocephalus \tab tay-sachs \tab reyes syndrome \tab unable to take care of self \tab other \tab don't know \tab no answer \tab Total \cr
   1990 \tab 720 \tab 1 \tab 84 \tab 53 \tab 17 \tab 7 \tab 20 \tab 29 \tab 44 \tab 104 \tab 90 \tab 22 \tab 16 \tab 51 \tab 9 \tab 2 \tab 4 \tab 6 \tab 14 \tab 6 \tab 17 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 8 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 7 \tab 6 \tab 11 \tab 1372 \cr
   Total \tab 720 \tab 1 \tab 84 \tab 53 \tab 17 \tab 7 \tab 20 \tab 29 \tab 44 \tab 104 \tab 90 \tab 22 \tab 16 \tab 51 \tab 9 \tab 2 \tab 4 \tab 6 \tab 14 \tab 6 \tab 17 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 8 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 7 \tab 6 \tab 11 \tab 1372 \cr
}
}

\seealso{
Other Genes: 
\code{\link{geneabrt}},
\code{\link{genedef1}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{geneself}}
}
\concept{Genes}
\keyword{variable}
