% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{genedef2}
\alias{genedef2}
\title{Meaning of serious genetic defect}
\source{
General Social Survey https://gss.norc.org
}
\description{
genedef2
}
\details{
Question When I say "serious genetic defect," what kinds of defects do you think of?  Any others?
RECORD VERBATIM AND CODE FIRST THREE MENTIONS
B. Second Mention
}
\section{Values}{

\itemize{
\item \code{1} downs syndrome
\item \code{2} mental retardation
\item \code{3} spina bifida
\item \code{4} muscular dystrophy
\item \code{5} cystic fibrosis
\item \code{6} cerebral palsy
\item \code{7} crippling diseases
\item \code{10} blind,deaf
\item \code{11} missing limbs
\item \code{12} major deformities
\item \code{13} life-threatening deform
\item \code{14} handicapped
\item \code{15} physical defects
\item \code{20} sickle cell anemia
\item \code{21} rh factor
\item \code{22} hemophilia
\item \code{30} addiction to drugs
\item \code{31} aids
\item \code{35} multiple sclerosis
\item \code{60} heart
\item \code{61} kidneys
\item \code{62} siamese twins
\item \code{70} incurable disease
\item \code{71} tourettes syndrome
\item \code{72} sexual transmitted diseases
\item \code{73} diabetes
\item \code{74} birth defects
\item \code{75} cancer
\item \code{76} psychological problems
\item \code{77} neurological problems
\item \code{78} hydrocephalus
\item \code{79} huntingtons chorea
\item \code{80} tay-sachs
\item \code{81} reyes syndrome
\item \code{90} unable to take care of self
\item \code{95} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5091/vshow}{GSS Data Explorer page for \code{genedef2}}.
}

\seealso{
Other Genes: 
\code{\link{geneabrt}},
\code{\link{genedef1}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{genegoo2}},
\code{\link{geneself}}
}
\concept{Genes}
\keyword{variable}
