% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{geneexps}
\alias{geneexps}
\title{Genes or experiences in determining personality}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneexps
}
\details{
Question For the following pair of questions, please circle the letter of the one that comes closer to your opinion:
}
\section{Values}{

\itemize{
\item \verb{[1]} genes play major role in determining personality
\item \verb{[2]} it is experience that determine personality
\item \verb{[9]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3046/vshow}{GSS Data Explorer page for \code{geneexps}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab genes play major role in determining personality \tab it is experience that determine personality \tab don't know \tab no answer \tab Total \cr
   2004 \tab 432 \tab 588 \tab 1723 \tab 51 \tab 18 \tab 2812 \cr
   Total \tab 432 \tab 588 \tab 1723 \tab 51 \tab 18 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{owndoing}},
\code{\link{sadblue}},
\code{\link{selfish}}
}
\concept{Genes and the Environment}
\keyword{variable}
