% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{geneexps}
\alias{geneexps}
\title{Genes or experiences in determining personality}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneexps
}
\details{
Question For the following pair of questions, please circle the letter of the one that comes closer to your opinion:
}
\section{Values}{

\itemize{
\item \code{1} genes play major role in determining personality
\item \code{2} it is experience that determine personality
\item \code{9} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3046/vshow}{GSS Data Explorer page for \code{geneexps}}.
}

\seealso{
Other Genes and the Environment: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{owndoing}},
\code{\link{sadblue}},
\code{\link{selfish}}
}
\concept{Genes and the Environment}
\concept{Genetic Testing}
\keyword{variable}
