% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{genegen1}
\alias{genegen1}
\title{How dangerous modifying genes in crops}
\source{
General Social Survey https://gss.norc.org
}
\description{
genegen1
}
\details{
Question And do you think that modifying the genes of certain crops is…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7493/vshow}{GSS Data Explorer page for \code{genegen1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 198 \tab 367 \tab 649 \tab 355 \tab 86 \tab 169 \tab 188 \tab 17 \tab 4032 \cr
   Total \tab 2003 \tab 198 \tab 367 \tab 649 \tab 355 \tab 86 \tab 169 \tab 188 \tab 17 \tab 4032 \cr
}
}

\keyword{variable}
