% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{genegets}
\alias{genegets}
\title{Employer vs. employee on access right to results}
\source{
General Social Survey https://gss.norc.org
}
\description{
genegets
}
\details{
Question C. Suppose a genetic screening test for certain forms of heart disease is made available by an employer to workers who want to take it. Who Do you think should have control over access to the test results -- the worker or the employer?
}
\section{Values}{

\itemize{
\item \code{1} employer should
\item \code{2} worker should
\item \code{3} both - voluntary
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1262/vshow}{GSS Data Explorer page for \code{genegets}}.
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}
}
\concept{Work Organizations}
\keyword{variable}
