% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{genegoo1}
\alias{genegoo1}
\title{Genetic testing good or harm}
\source{
General Social Survey https://gss.norc.org
}
\description{
genegoo1
}
\details{
Question B. Some people say that genetic screening may cause trouble.
Others think it is a wonderful medical advance. Based on what you know, do you think genetic screening will do more harm than good or more good than harm?
}
\section{Values}{

\itemize{
\item \verb{[1]} harm ) good
\item \verb{[2]} good ) harm
\item \verb{[7]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5086/vshow}{GSS Data Explorer page for \code{genegoo1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab harm ) good \tab good ) harm \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2431 \tab 107 \tab 235 \tab 103 \tab 28 \tab 2904 \cr
   Total \tab 2431 \tab 107 \tab 235 \tab 103 \tab 28 \tab 2904 \cr
}
}

\seealso{
Other Genes: 
\code{\link{geneabrt}},
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo2}},
\code{\link{geneself}}
}
\concept{Genes}
\keyword{variable}
