% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{genegoo2}
\alias{genegoo2}
\title{Genetic testing good or harm}
\source{
General Social Survey https://gss.norc.org
}
\description{
genegoo2
}
\details{
Question C. Some people say that genetic testing may cause trouble. Others think it is a wonderful medical advance. Based on what you know, do you think genetic testing will do more harm than good or more good than harm?
}
\section{Values}{

\itemize{
\item \code{1} more good than harm
\item \code{2} more harm than good
\item \code{3} it depends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2018 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5087/vshow}{GSS Data Explorer page for \code{genegoo2}}.
}

\seealso{
Other Genes: 
\code{\link{geneabrt}},
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{geneself}}

Other Genetic Testing: 
\code{\link{geneabrt}},
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{geneself}}
}
\concept{Genes}
\concept{Genetic Testing}
\keyword{variable}
