% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{genegood}
\alias{genegood}
\title{Genetic testing good or harm}
\source{
General Social Survey https://gss.norc.org
}
\description{
genegood
}
\details{
Question A. Some people say that genetic screening is a wonderful medical advance. Others think it may cause trouble. Based on what you know, do you think genetic screening will do more good than harm, or more harm than good?
}
\section{Values}{

\itemize{
\item \code{1} good ) harm
\item \code{2} harm ) good
\item \code{3} it depends
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 1990, 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5085/vshow}{GSS Data Explorer page for \code{genegood}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab good ) harm \tab harm ) good \tab it depends \tab refused \tab don't know \tab no answer \tab Total \cr
   1990 \tab 455 \tab 434 \tab 190 \tab 99 \tab 6 \tab 183 \tab 5 \tab 1372 \cr
   1996 \tab 2408 \tab 242 \tab 112 \tab - \tab - \tab 119 \tab 23 \tab 2904 \cr
   Total \tab 2863 \tab 676 \tab 302 \tab 99 \tab 6 \tab 302 \tab 28 \tab 4276 \cr
}
}

\seealso{
Other Genes: 
\code{\link{genetest}},
\code{\link{genetst1}}
}
\concept{Genes}
\keyword{variable}
