% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_134.R
\name{genehire}
\alias{genehire}
\title{Employers' right to hire by genetic tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
genehire
}
\details{
Question B. Should employers have the right not to hire workers if tests show they have an inherited tendency to develop certain forms of cancer or heart disease, or should they not have the right?
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1259/vshow}{GSS Data Explorer page for \code{genehire}}.
}

\seealso{
Other Work Organizations: 
\code{\link{genejob}}
}
\concept{Work Organizations}
\keyword{variable}
