% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_134.R
\name{genejob}
\alias{genejob}
\title{Should employers give genetic tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
genejob
}
\details{
Question Here are some questions about a new scientific technology called "genetic screening."  Using genetic screening, it is now possible to tell whether someone has inherited a tendency to develop certain cancers and certain forms of heart disease. These tests do not mean that a person will always develop the disease, but only that he or she may do so, depending on other conditions.
A. Should employers have the right to give these genetic tests to people who are applying for a job, or shouldn't they have that right?
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1258/vshow}{GSS Data Explorer page for \code{genejob}}.
}

\seealso{
Other Work Organizations: 
\code{\link{genehire}}

Other Genetic Testing: 
\code{\link{genehire}}
}
\concept{Genetic Testing}
\concept{Work Organizations}
\keyword{variable}
