% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_339.R
\name{genenvo3}
\alias{genenvo3}
\title{Gene or learning?}
\source{
General Social Survey https://gss.norc.org
}
\description{
genenvo3
}
\details{
Question (The description of the third person presented to the respondent - please check GENEVIG1 & GENEVIG2)(Please type in a number from 1 to 21):
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3034/vshow}{GSS Data Explorer page for \code{genenvo3}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 432 \tab 155 \tab 239 \tab 185 \tab 153 \tab 45 \tab 64 \tab 203 \tab 78 \tab 98 \tab 147 \tab 73 \tab 43 \tab 93 \tab 296 \tab 50 \tab 57 \tab 144 \tab 44 \tab 55 \tab 66 \tab 60 \tab 18 \tab 14 \tab 2812 \cr
   Total \tab 432 \tab 155 \tab 239 \tab 185 \tab 153 \tab 45 \tab 64 \tab 203 \tab 78 \tab 98 \tab 147 \tab 73 \tab 43 \tab 93 \tab 296 \tab 50 \tab 57 \tab 144 \tab 44 \tab 55 \tab 66 \tab 60 \tab 18 \tab 14 \tab 2812 \cr
}
}

\seealso{
Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{genenvo4}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{pessimst}},
\code{\link{satself}},
\code{\link{slfrspct}}
}
\concept{Genes and the Environment}
\keyword{variable}
