% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_339.R
\name{genenvo3}
\alias{genenvo3}
\title{Gene or learning?}
\source{
General Social Survey https://gss.norc.org
}
\description{
genenvo3
}
\details{
Question (The description of the third person presented to the respondent - please check GENEVIG1 & GENEVIG2)(Please type in a number from 1 to 21):
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3034/vshow}{GSS Data Explorer page for \code{genenvo3}}.
}

\seealso{
Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genenvo4}},
\code{\link{genevig1}},
\code{\link{genevig2}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{owndoing}},
\code{\link{pessimst}},
\code{\link{sadblue}},
\code{\link{satself}},
\code{\link{selfish}},
\code{\link{slfrspct}}
}
\concept{Genes and the Environment}
\keyword{variable}
