% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{genes}
\alias{genes}
\title{Traits are transferred over generation}
\source{
General Social Survey https://gss.norc.org
}
\description{
genes
}
\details{
Question Traits are transferred from generation to generation through the…
}
\section{Values}{

\itemize{
\item \code{1} sperm only
\item \code{2} egg only
\item \code{3} sperm and egg
\item \code{4} testes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3479/vshow}{GSS Data Explorer page for \code{genes}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
