% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{geneself}
\alias{geneself}
\title{Would genetically test fetus}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneself
}
\details{
Question Today, tests are being developed that make it possible to detect serious genetic defects before a baby is born. But so far, it is impossible either to treat or to correct most of them. If (you/your partner) were pregnant, would you want (her) to have a test to find out if the baby has any serious genetic defects?
}
\section{Values}{

\itemize{
\item \code{1} have test
\item \code{2} not have test
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/C/- \tab partial \cr
   1996 \tab A/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
   2018 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5088/vshow}{GSS Data Explorer page for \code{geneself}}.
}

\seealso{
Other Genes: 
\code{\link{geneabrt}},
\code{\link{genedef1}},
\code{\link{genedef2}},
\code{\link{genedef3}},
\code{\link{genegoo1}},
\code{\link{genegoo2}}
}
\concept{Genes}
\keyword{variable}
