% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{geneself2}
\alias{geneself2}
\title{R would want genetic test done on fetus}
\source{
General Social Survey https://gss.norc.org
}
\description{
geneself2
}
\details{
Question Today, tests are being developed that make it possible to detect serious genetic defects before a baby is born. But so far, it is impossible either to treat or to correct most of them. If (you were pregnant/you were expecting a child/your partner were pregnant), would you want (the pregnant woman/her) to have a test to find out if the baby has any serious genetic defects?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7295/vshow}{GSS Data Explorer page for \code{geneself2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1168 \tab 740 \tab 389 \tab 21 \tab 30 \tab 2348 \cr
   Total \tab 1168 \tab 740 \tab 389 \tab 21 \tab 30 \tab 2348 \cr
}
}

\keyword{variable}
