% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{genetest}
\alias{genetest}
\title{Know about genetic testing}
\source{
General Social Survey https://gss.norc.org
}
\description{
genetest
}
\details{
Question A. How much would you say you have heard or read about genetic screening - a great deal, something but not very much, or nothing at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} not very much
\item \verb{[3]} nothing at all
\item \verb{[7]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 1990, 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5083/vshow}{GSS Data Explorer page for \code{genetest}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab not very much \tab nothing at all \tab refused \tab don't know \tab no answer \tab Total \cr
   1990 \tab 455 \tab 117 \tab 541 \tab 237 \tab 2 \tab 16 \tab 4 \tab 1372 \cr
   1996 \tab 2408 \tab 67 \tab 313 \tab 86 \tab - \tab 11 \tab 19 \tab 2904 \cr
   Total \tab 2863 \tab 184 \tab 854 \tab 323 \tab 2 \tab 27 \tab 23 \tab 4276 \cr
}
}

\seealso{
Other Genes: 
\code{\link{genegood}},
\code{\link{genetst1}}
}
\concept{Genes}
\keyword{variable}
