% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{genetest}
\alias{genetest}
\title{Know about genetic testing}
\source{
General Social Survey https://gss.norc.org
}
\description{
genetest
}
\details{
Question A. How much would you say you have heard or read about genetic screening - a great deal, something but not very much, or nothing at all?
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} not very much
\item \code{3} nothing at all
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5083/vshow}{GSS Data Explorer page for \code{genetest}}.
}

\seealso{
Other Genes: 
\code{\link{genegood}},
\code{\link{genetst1}}

Other Genetic Testing: 
\code{\link{genegood}},
\code{\link{genetst1}}
}
\concept{Genes}
\concept{Genetic Testing}
\keyword{variable}
