% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{genetics}
\alias{genetics}
\title{How likely x s situation caused by genetic problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
genetics
}
\details{
Question In your opinion, how likely is it that NAME's situation might be caused by (READ a-f) -- very likely, somewhat likely, not very likely, or not at all likely?
E. A genetic or inherited problem
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1992/vshow}{GSS Data Explorer page for \code{genetics}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 155 \tab 507 \tab 419 \tab 269 \tab 78 \tab 16 \tab 2904 \cr
   2006 \tab 3073 \tab 201 \tab 681 \tab 331 \tab 152 \tab 58 \tab 14 \tab 4510 \cr
   2018 \tab 1175 \tab 167 \tab 498 \tab 288 \tab 163 \tab 28 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 523 \tab 1686 \tab 1038 \tab 584 \tab 164 \tab 59 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
