% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{genevig1}
\alias{genevig1}
\title{Vignette number used}
\source{
General Social Survey https://gss.norc.org
}
\description{
genevig1
}
\details{
Question Vignette set number given to respondent.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3030/vshow}{GSS Data Explorer page for \code{genevig1}}.
}

\seealso{
Other Genes and the Environment: 
\code{\link{genenvo1}},
\code{\link{genenvo2}},
\code{\link{genevig2}}
}
\concept{Genes and the Environment}
\keyword{variable}
