% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{geomobil}
\alias{geomobil}
\title{Has r moved from one kind of place to another}
\source{
General Social Survey https://gss.norc.org
}
\description{
geomobil
}
\details{
Question We are interested in the extent to which you have moved from one kind of place to another. Which do you think is most like your experience of life? Which of the following best applies to you?
}
\section{Values}{

\itemize{
\item \verb{[1]} i have lived in different countries
\item \verb{[2]} i have lived in different places in the same country
\item \verb{[3]} i have lived in different neighborhoods in the same place
\item \verb{[4]} i have always lived in the same neighborhood
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4698/vshow}{GSS Data Explorer page for \code{geomobil}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab i have lived in different countries \tab i have lived in different places in the same country \tab i have lived in different neighborhoods in the same place \tab i have always lived in the same neighborhood \tab don't know \tab Total \cr
   2008 \tab 658 \tab 205 \tab 613 \tab 385 \tab 161 \tab 1 \tab 2023 \cr
   Total \tab 658 \tab 205 \tab 613 \tab 385 \tab 161 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{ISSP Religion}
\keyword{variable}
