% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{geomobil}
\alias{geomobil}
\title{Has r moved from one kind of place to another}
\source{
General Social Survey https://gss.norc.org
}
\description{
geomobil
}
\details{
Question We are interested in the extent to which you have moved from one kind of place to another. Which do you think is most like your experience of life? Which of the following best applies to you?
}
\section{Values}{

\itemize{
\item \code{1} i have lived in different countries
\item \code{2} i have lived in different places in the same country
\item \code{3} i have lived in different neighborhoods in the same place
\item \code{4} i have always lived in the same neighborhood
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4698/vshow}{GSS Data Explorer page for \code{geomobil}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{Dwelling (R's)}
\concept{ISSP Religion}
\concept{Residence}
\keyword{variable}
