% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_661.R
\name{gestate}
\alias{gestate}
\title{Abortion legal past 4.5 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
gestate
}
\details{
Question Please tell us whether or not you think it should be possible for a pregnant woman to obtain a legal abortion if... ...she is beyond 4.5 months, that is, about halfway through the pregnancy.

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7945/vshow}{GSS Data Explorer page for \code{gestate}}.
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16frndnum}},
\code{\link{brv16gpanum}},
\code{\link{brv16othnum}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}}
}
\concept{GSS Next}
\keyword{variable}
