% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_393.R
\name{gesttalk}
\alias{gesttalk}
\title{How often r see people making strange gesture in public}
\source{
General Social Survey https://gss.norc.org
}
\description{
gesttalk
}
\details{
Question How often do you see people in public who make strange gestures or talk when there was no one there to listen…
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} almost never
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3569/vshow}{GSS Data Explorer page for \code{gesttalk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab almost never \tab never \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3073 \tab 268 \tab 540 \tab 438 \tab 165 \tab 13 \tab 13 \tab 4510 \cr
   Total \tab 3073 \tab 268 \tab 540 \tab 438 \tab 165 \tab 13 \tab 13 \tab 4510 \cr
}
}

\seealso{
Other Mental Health III: 
\code{\link{avoidmh}}
}
\concept{Mental Health III}
\keyword{variable}
