% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_393.R
\name{gesttalk}
\alias{gesttalk}
\title{How often r see people making strange gesture in public}
\source{
General Social Survey https://gss.norc.org
}
\description{
gesttalk
}
\details{
Question How often do you see people in public who make strange gestures or talk when there was no one there to listen…
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} almost never
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3569/vshow}{GSS Data Explorer page for \code{gesttalk}}.
}

\seealso{
Other Mental Health III: 
\code{\link{avoidmh}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
