% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_407.R
\name{getaheay}
\alias{getaheay}
\title{Opinion of why people get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
getaheay
}
\details{
Question Which do you think is the most important reason why people get ahead: their own hard work or lucky breaks and help from other people?
}
\section{Values}{

\itemize{
\item \code{1} hard work most important
\item \code{2} hard work, luck equally important
\item \code{3} luck or help from other people most important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3711/vshow}{GSS Data Explorer page for \code{getaheay}}.
}

\seealso{
Other Misc I: 
\code{\link{natdiscm}},
\code{\link{newsprnt}}
}
\concept{Misc I}
\keyword{variable}
