% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{getjob}
\alias{getjob}
\title{How likely you would find a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
getjob
}
\details{
Question How likely do you think it is that you would find a job?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} likely
\item \verb{[3]} unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4597/vshow}{GSS Data Explorer page for \code{getjob}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab likely \tab unlikely \tab very unlikely \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 2460 \tab 80 \tab 82 \tab 37 \tab 109 \tab 29 \tab 35 \tab 2832 \cr
   2006 \tab 4008 \tab 151 \tab 111 \tab 81 \tab 143 \tab 12 \tab 4 \tab 4510 \cr
   2016 \tab 2533 \tab 94 \tab 97 \tab 58 \tab 60 \tab 2 \tab 23 \tab 2867 \cr
   Total \tab 9001 \tab 325 \tab 290 \tab 176 \tab 312 \tab 43 \tab 62 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
