% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{gills}
\alias{gills}
\title{How fish get oxygen}
\source{
General Social Survey https://gss.norc.org
}
\description{
gills
}
\details{
Question How do most fish get the oxygen they need to survive?
}
\section{Values}{

\itemize{
\item \verb{[1]} they take in water and break it down into hydrogen and oxygen
\item \verb{[2]} using their gills, they take in oxygen that is dissolved in water
\item \verb{[3]} they get their oxygen from the food they eat
\item \verb{[4]} they come to the surface every few minutes to breathe air into their lungs
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3480/vshow}{GSS Data Explorer page for \code{gills}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab they take in water and break it down into hydrogen and oxygen \tab using their gills, they take in oxygen that is dissolved in water \tab they get their oxygen from the food they eat \tab they come to the surface every few minutes to breathe air into their lungs \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 137 \tab 1124 \tab 29 \tab 87 \tab 107 \tab 21 \tab 2023 \cr
   Total \tab 518 \tab 137 \tab 1124 \tab 29 \tab 87 \tab 107 \tab 21 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{seesand}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
