% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{givearts}
\alias{givearts}
\title{Contributions to arts culture humanities}
\source{
General Social Survey https://gss.norc.org
}
\description{
givearts
}
\details{
Question During the last year, approximately how much money did you and the other family members in your household contribute to each of the following:
D. Were any of your contributions to organizations involved in the arts, culture, or the humanities?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1142/vshow}{GSS Data Explorer page for \code{givearts}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2022 \tab 307 \tab 405 \tab 3 \tab 95 \tab 2832 \cr
   Total \tab 2022 \tab 307 \tab 405 \tab 3 \tab 95 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{kid5up}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}},
\code{\link{relhrs2}}
}
\concept{Religion}
\keyword{variable}
