% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{givearts}
\alias{givearts}
\title{Contributions to arts culture humanities}
\source{
General Social Survey https://gss.norc.org
}
\description{
givearts
}
\details{
Question During the last year, approximately how much money did you and the other family members in your household contribute to each of the following:
D. Were any of your contributions to organizations involved in the arts, culture, or the humanities?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1142/vshow}{GSS Data Explorer page for \code{givearts}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2022 \tab 307 \tab 405 \tab 3 \tab 95 \tab 2832 \cr
   Total \tab 2022 \tab 307 \tab 405 \tab 3 \tab 95 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{kid5up}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}},
\code{\link{relhrs2}}
}
\concept{Religion}
\keyword{variable}
