% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_471.R
\name{givexec}
\alias{givexec}
\title{How much ought corporate head make?}
\source{
General Social Survey https://gss.norc.org
}
\description{
givexec
}
\details{
Question Next, what do you think people in these jobs ought to be paid -- how much do you think they should earn each year  before taxes, regardless of what they actually get? Please write in how much they should earn each year.
E. The chairman of a large national corporation?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999992} more than now paid
\item \code{99999993} same as now paid
\item \code{99999994} less than now paid
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4283/vshow}{GSS Data Explorer page for \code{givexec}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10000 \tab 105000 \tab 11000 \tab 110000 \tab 12000 \tab 120000 \tab 12500 \tab 125000 \tab 135000 \tab 14000 \tab 15000 \tab 150000 \tab 16000 \tab 170000 \tab 175000 \tab 18000 \tab 189000 \tab 1e+05 \tab $1 million (or more, for 1987 & 2000) \tab 200 \tab 2000 \tab 20000 \tab 21000 \tab 22000 \tab 220000 \tab 23000 \tab 230000 \tab 24000 \tab 245000 \tab 250 \tab 25000 \tab 250000 \tab 250875 \tab 27000 \tab 275000 \tab 28000 \tab 28800 \tab 29000 \tab 2e+05 \tab 3000 \tab 30000 \tab 31200 \tab 34000 \tab 35000 \tab 350000 \tab 38000 \tab 3e+05 \tab 4000 \tab 40000 \tab 42000 \tab 45000 \tab 450000 \tab 46000 \tab 47000 \tab 4e+05 \tab 5000 \tab 50000 \tab 52000 \tab 54000 \tab 55000 \tab 5830 \tab 5e+05 \tab 6000 \tab 60000 \tab 65000 \tab 68000 \tab 6e+05 \tab 7000 \tab 70000 \tab 72800 \tab 75000 \tab 750000 \tab 7800 \tab 7e+05 \tab 8000 \tab 80000 \tab 81000 \tab 85000 \tab 850000 \tab 8e+05 \tab 9000 \tab 90000 \tab 95000 \tab more than now paid \tab same as now paid \tab less than now paid \tab 9e+05 \tab can't choose \tab no answer \tab 1300 \tab 130000 \tab 140000 \tab 160000 \tab 190000 \tab 2250 \tab 225000 \tab 325000 \tab 39000 \tab 48000 \tab 57000 \tab 78000 \tab 89000 \tab 0 \tab 1 \tab 1.3e+07 \tab 1.5e+07 \tab 100 \tab 1000 \tab 115000 \tab 1200000 \tab 133000 \tab 13577 \tab 15 \tab 150 \tab 1500000 \tab 180000 \tab 1900000 \tab 1e+07 \tab 2 \tab 2.5 \tab 2.5e+07 \tab 2100 \tab 210000 \tab 22 \tab 23333 \tab 25 \tab 2500000 \tab 254869 \tab 2e+06 \tab 2e+07 \tab 3000009 \tab 32131 \tab 333333 \tab 340000 \tab 35 \tab 3500000 \tab 35350 \tab 360000 \tab 361 \tab 366 \tab 375000 \tab 395000 \tab 3e+06 \tab 3e+07 \tab 46729 \tab 4e+06 \tab 5 \tab 50 \tab 500 \tab 50900 \tab 525000 \tab 550000 \tab 5e+06 \tab 5e+07 \tab 626262 \tab 65777 \tab 6776900 \tab 6e+06 \tab 7.5e+07 \tab 7e+06 \tab 85 \tab 86 \tab 8888886 \tab 8e+06 \tab 99000 \tab 9918 \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 10 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 1 \tab 7 \tab 1 \tab 1 \tab 7 \tab 79 \tab 5 \tab 1 \tab 7 \tab 2 \tab 1 \tab 191 \tab 19 \tab 1 \tab 1 \tab 27 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 33 \tab 37 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 66 \tab 1 \tab 52 \tab 2 \tab 1 \tab 26 \tab 5 \tab 2 \tab 25 \tab 1 \tab 65 \tab 2 \tab 14 \tab 3 \tab 1 \tab 1 \tab 8 \tab 3 \tab 133 \tab 2 \tab 1 \tab 10 \tab 1 \tab 44 \tab 2 \tab 64 \tab 11 \tab 1 \tab 5 \tab 1 \tab 32 \tab 1 \tab 76 \tab 5 \tab 1 \tab 3 \tab 1 \tab 63 \tab 1 \tab 7 \tab 2 \tab 5 \tab 2 \tab 23 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 81 \tab 234 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 18 \tab - \tab 10 \tab - \tab - \tab 5 \tab 58 \tab - \tab 2 \tab 3 \tab - \tab - \tab 174 \tab 82 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 8 \tab 52 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 101 \tab 1 \tab 11 \tab 1 \tab - \tab 5 \tab 7 \tab - \tab 34 \tab - \tab 18 \tab - \tab 2 \tab 2 \tab - \tab - \tab 14 \tab - \tab 61 \tab - \tab 1 \tab 3 \tab - \tab 66 \tab - \tab 55 \tab 8 \tab - \tab 6 \tab - \tab 28 \tab - \tab 46 \tab 9 \tab - \tab 1 \tab - \tab 50 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 14 \tab 1 \tab - \tab - \tab - \tab 3 \tab 70 \tab 204 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 14 \tab 1 \tab 1 \tab 2 \tab 1 \tab 12 \tab 1 \tab 15 \tab 1 \tab - \tab 1 \tab 115 \tab - \tab 2 \tab 12 \tab - \tab - \tab 173 \tab 170 \tab 2 \tab 2 \tab 10 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 4 \tab 106 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 191 \tab - \tab 5 \tab - \tab - \tab 1 \tab 16 \tab - \tab 104 \tab - \tab 4 \tab - \tab 3 \tab 5 \tab - \tab - \tab 37 \tab 3 \tab 32 \tab - \tab 1 \tab 1 \tab - \tab 224 \tab - \tab 26 \tab 1 \tab - \tab 12 \tab 2 \tab 11 \tab - \tab 22 \tab 26 \tab - \tab 11 \tab - \tab 41 \tab - \tab 5 \tab 3 \tab 12 \tab - \tab 15 \tab 2 \tab - \tab - \tab - \tab 3 \tab 21 \tab 133 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 6 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 17 \tab 2 \tab 1 \tab 33 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 15 \tab 1 \tab 61 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 19 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 52 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 87 \tab 4032 \cr
   Total \tab 3829 \tab 27 \tab 3 \tab 2 \tab 6 \tab 5 \tab 36 \tab 2 \tab 32 \tab 2 \tab 1 \tab 13 \tab 252 \tab 5 \tab 5 \tab 22 \tab 2 \tab 1 \tab 538 \tab 271 \tab 3 \tab 4 \tab 41 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 45 \tab 195 \tab 1 \tab 4 \tab 2 \tab 3 \tab 1 \tab 1 \tab 358 \tab 2 \tab 68 \tab 3 \tab 1 \tab 32 \tab 28 \tab 2 \tab 163 \tab 1 \tab 87 \tab 2 \tab 19 \tab 10 \tab 1 \tab 1 \tab 59 \tab 6 \tab 226 \tab 2 \tab 3 \tab 14 \tab 1 \tab 334 \tab 2 \tab 145 \tab 20 \tab 1 \tab 23 \tab 3 \tab 71 \tab 1 \tab 144 \tab 40 \tab 1 \tab 15 \tab 1 \tab 154 \tab 1 \tab 16 \tab 6 \tab 20 \tab 2 \tab 52 \tab 6 \tab 1 \tab 1 \tab 3 \tab 8 \tab 172 \tab 571 \tab 1 \tab 6 \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 17 \tab 2 \tab 1 \tab 33 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 15 \tab 1 \tab 61 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 19 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 52 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 87 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givbus}},
\code{\link{givcabnt}},
\code{\link{givfarm}},
\code{\link{givlaw}},
\code{\link{givowner}},
\code{\link{givsales}},
\code{\link{givsec}},
\code{\link{givskill}},
\code{\link{givunskl}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
