% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_472.R
\name{givfctry}
\alias{givfctry}
\title{How much ought owner of a large factory makes}
\source{
General Social Survey https://gss.norc.org
}
\description{
givfctry
}
\details{
Question Next, what do you think people in these jobs ought to be paid -- how much do you think they should earn each year  before taxes, regardless of what they actually get? Please write in how much they should earn each year.
N. The owner-manager of a large factory?
}
\section{Values}{

\itemize{
\item \verb{[1e+06]} $1 million (or more, for 1987 & 2000)
\item \verb{[99999992]} more than now paid
\item \verb{[99999993]} same as now paid
\item \verb{[99999994]} less than now paid
\item \verb{[99999997]} never worked
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4292/vshow}{GSS Data Explorer page for \code{givfctry}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab 55000 \tab Total \cr
   2000 \tab 2816 \tab 1 \tab 2817 \cr
   Total \tab 2816 \tab 1 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givjudge}},
\code{\link{givrocc}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incgap}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
