% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_472.R
\name{givjudge}
\alias{givjudge}
\title{How much ought justice make}
\source{
General Social Survey https://gss.norc.org
}
\description{
givjudge
}
\details{
Question Next, what do you think people in these jobs ought to be paid -- how much do you think they should earn each year  before taxes, regardless of what they actually get? Please write in how much they should earn each year.
O. A Supreme Court Justice?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999992} more than now paid
\item \code{99999993} same as now paid
\item \code{99999994} less than now paid
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4293/vshow}{GSS Data Explorer page for \code{givjudge}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givfctry}},
\code{\link{givrocc}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incgap}}

Other Inequality, Income: 
\code{\link{givfctry}},
\code{\link{givrocc}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}},
\code{\link{incgap}}

Other Occupation: 
\code{\link{givfctry}},
\code{\link{givrocc}}
}
\concept{ISSP Social Inequality}
\concept{Inequality, Income}
\concept{Occupation}
\keyword{variable}
