% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_206.R
\name{givoth}
\alias{givoth}
\title{Contributed money in other (specify)}
\source{
General Social Survey https://gss.norc.org
}
\description{
givoth
}
\details{
Question Listed on this card are examples of many different fields in which people and families contribute money or other property for charitable purposes. I mean making a voluntary contribution and not with the intention of making a profit or obtaining goods and/or services for yourself. In which, if any, of the fields listed on this card have you and the members of your family or household contributed some money or other property in 1995? Just read off the letter of each field (DON'T READ. RECORD BELOW UNDER GIVHLTH-GIVOTH.)
O. Other (SPECIFY)
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1944/vshow}{GSS Data Explorer page for \code{givoth}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{givinfrm}},
\code{\link{givintl}},
\code{\link{toteduc}},
\code{\link{totenvir}},
\code{\link{tothlth}},
\code{\link{tothuman}},
\code{\link{totpub}},
\code{\link{totrec}},
\code{\link{totrelig}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
