% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{gochurch}
\alias{gochurch}
\title{Importance of attending church regularly}
\source{
General Social Survey https://gss.norc.org
}
\description{
gochurch
}
\details{
Question People have many differing views about what makes a person a good Christian or Jew. Please tell me how important is each of the following to you. Once again, please indicate where you would place your feelings on a scale from 1 to 5. How important is it …
A. To attend regularly religious services at church or synagogue.
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{5} not very important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1089/vshow}{GSS Data Explorer page for \code{gochurch}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab 2 \tab 3 \tab 4 \tab not very important \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 418 \tab 218 \tab 298 \tab 205 \tab 318 \tab 13 \tab 11 \tab 1481 \cr
   1998 \tab 1387 \tab 428 \tab 183 \tab 292 \tab 158 \tab 345 \tab 27 \tab 12 \tab 2832 \cr
   Total \tab 1387 \tab 846 \tab 401 \tab 590 \tab 363 \tab 663 \tab 40 \tab 23 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{blkwhite}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{permoral}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{rotapple}}
}
\concept{Religion}
\keyword{variable}
