% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_229.R
\name{godlove}
\alias{godlove}
\title{Feel gods love directly or thru others}
\source{
General Social Survey https://gss.norc.org
}
\description{
godlove
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
E. I feel God's love for me, directly or through others.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never/almst nevr
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2112/vshow}{GSS Data Explorer page for \code{godlove}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{beausprt}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{drama}},
\code{\link{effort}},
\code{\link{feelgod}},
\code{\link{fidgety}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{lesspain}},
\code{\link{localart}},
\code{\link{meditate}},
\code{\link{natarts}},
\code{\link{natlart}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{popmusic}},
\code{\link{privpray}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{rellife}},
\code{\link{relpersn}},
\code{\link{sprtprsn}},
\code{\link{stateart}},
\code{\link{uniongod}},
\code{\link{volarts}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
