% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_344.R
\name{godlvoth}
\alias{godlvoth}
\title{Feel gods love for me through others}
\source{
General Social Survey https://gss.norc.org
}
\description{
godlvoth
}
\details{
Question I feel God''s love for me, through others.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never or almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3090/vshow}{GSS Data Explorer page for \code{godlvoth}}.
}

\seealso{
Other Spirituality and Religion: 
\code{\link{blessngs}},
\code{\link{closrgod}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlvdir}},
\code{\link{joylifts}},
\code{\link{relsprt1}},
\code{\link{relsprt2}}

Other Religion: 
\code{\link{blessngs}},
\code{\link{closrgod}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlvdir}},
\code{\link{joylifts}},
\code{\link{relsprt1}},
\code{\link{relsprt2}}
}
\concept{Religion}
\concept{Spirituality and Religion}
\keyword{variable}
