% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{godparts}
\alias{godparts}
\title{How often does r contact god-parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
godparts
}
\details{
Question Now some questions about your contact with other relatives. Please indicate how often you have been in contact with any of the following types of relatives in the last four weeks.
F. God-parents
}
\section{Values}{

\itemize{
\item \code{1} more than twice in last 4 weeks
\item \code{2} once or twice in last 4 weeks
\item \code{3} not at all in last 4 weeks
\item \code{4} i have no living relative of this type
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4201/vshow}{GSS Data Explorer page for \code{godparts}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
