% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{godsells}
\alias{godsells}
\title{Religious image to sell products}
\source{
General Social Survey https://gss.norc.org
}
\description{
godsells
}
\details{
Question Do you strongly approve of, somewhat approve of, neither approve of nor oppose, somewhat oppose, or strongly oppose the following public expressions of religion:
A. The use of religious "images" in public advertising to sell non-religious commercial products?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly approve
\item \verb{[2]} somewhat approve
\item \verb{[3]} not apprve/oppse
\item \verb{[4]} somewhat oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1137/vshow}{GSS Data Explorer page for \code{godsells}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly approve \tab somewhat approve \tab not apprve/oppse \tab somewhat oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 35 \tab 79 \tab 320 \tab 314 \tab 619 \tab 68 \tab 10 \tab 2832 \cr
   Total \tab 1387 \tab 35 \tab 79 \tab 320 \tab 314 \tab 619 \tab 68 \tab 10 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsport}},
\code{\link{numbwfrd}},
\code{\link{numfrend}}
}
\concept{Religion}
\keyword{variable}
