% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{goldfish}
\alias{goldfish}
\title{Experiment about goldfish and temperature}
\source{
General Social Survey https://gss.norc.org
}
\description{
goldfish
}
\details{
Question Experiment about goldfish and temperature.
}
\section{Values}{

\itemize{
\item \code{1} experiment a
\item \code{2} experiment b
\item \code{3} experiment c
\item \code{4} experiment d
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3476/vshow}{GSS Data Explorer page for \code{goldfish}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab experiment a \tab experiment b \tab experiment c \tab experiment d \tab don't know \tab no answer \tab Total \cr
   2008 \tab 709 \tab 761 \tab 84 \tab 116 \tab 77 \tab 248 \tab 28 \tab 2023 \cr
   Total \tab 709 \tab 761 \tab 84 \tab 116 \tab 77 \tab 248 \tab 28 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
