% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_286.R
\name{googlesn}
\alias{googlesn}
\title{Use googlesn}
\source{
General Social Survey https://gss.norc.org
}
\description{
googlesn
}
\details{
Question (Which of the following social networking or social media sites are you a member or regular user of?)
Google+ -  Google's social network platform.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6276/vshow}{GSS Data Explorer page for \code{googlesn}}.
}

\seealso{
Other Social Media Usage: 
\code{\link{clssmtes}},
\code{\link{facebook}},
\code{\link{flickr}},
\code{\link{instagrm}},
\code{\link{intmbile}},
\code{\link{intrecnt}},
\code{\link{intuse}},
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{intwkenm}},
\code{\link{linkedin}},
\code{\link{pinterst}},
\code{\link{snapchat}},
\code{\link{snsmfrst}},
\code{\link{snsmnew}},
\code{\link{snsmot2a}},
\code{\link{snsmot2b}},
\code{\link{snsmot2c}},
\code{\link{snsmoth1}},
\code{\link{tumblr}},
\code{\link{twitter}},
\code{\link{vine}},
\code{\link{whatsapp}}
}
\concept{Social Media Usage}
\keyword{variable}
