% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{govaid}
\alias{govaid}
\title{Ever receive welfare, unemp insur, etc.}
\source{
General Social Survey https://gss.norc.org
}
\description{
govaid
}
\details{
Question Did you ever--because of sickness, unemployment, or any other reason--receive anything like welfare, unemployment insurance, or other aid from government agencies?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab - \tab full \cr
   1974 \tab - \tab full \cr
   1975 \tab - \tab full \cr
   1976 \tab - \tab full \cr
   1977 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/582/vshow}{GSS Data Explorer page for \code{govaid}}.
}

\seealso{
Other Core: 
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{parsol}},
\code{\link{unemp}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}}

Other Income: 
\code{\link{getaid}}

Other Welfare: 
\code{\link{getaid}}
}
\concept{Core}
\concept{Income}
\concept{Welfare}
\keyword{variable}
