% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{govcare}
\alias{govcare}
\title{Govt concern for social welfare}
\source{
General Social Survey https://gss.norc.org
}
\description{
govcare
}
\details{
Question B. For some time there have been discussions about how concerned the government should be about the social welfare of people. We have listed here two different opinions. With which opinion do you agree the most?  READ STATEMENTS A & B. CIRCLE ONE CODE.
}
\section{Values}{

\itemize{
\item \verb{[1]} too much care
\item \verb{[2]} on right track
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/242/vshow}{GSS Data Explorer page for \code{govcare}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab too much care \tab on right track \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 367 \tab 1013 \tab 75 \tab 18 \tab 1473 \cr
   Total \tab 0 \tab 367 \tab 1013 \tab 75 \tab 18 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}},
\code{\link{tax}}
}
\concept{Core}
\keyword{variable}
