% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{govcheat}
\alias{govcheat}
\title{Wrong to give govt false info for benefits?}
\source{
General Social Survey https://gss.norc.org
}
\description{
govcheat
}
\details{
Question Consider the situations listed below. Do you feel it is wrong or not wrong if. . .
B. A person gives the government incorrect information about himself to get government benefits that he is not entitled to.
}
\section{Values}{

\itemize{
\item \code{1} not wrong
\item \code{2} a bit wrong
\item \code{3} wrong
\item \code{4} seriously wrong
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4638/vshow}{GSS Data Explorer page for \code{govcheat}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not wrong \tab a bit wrong \tab wrong \tab seriously wrong \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 27 \tab 49 \tab 586 \tab 645 \tab 25 \tab 27 \tab 1517 \cr
   1998 \tab 1548 \tab 22 \tab 34 \tab 510 \tab 664 \tab 27 \tab 27 \tab 2832 \cr
   Total \tab 1706 \tab 49 \tab 83 \tab 1096 \tab 1309 \tab 52 \tab 54 \tab 4349 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{homosex1}},
\code{\link{taxcheat}}
}
\concept{ISSP Religion}
\keyword{variable}
