% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{governor}
\alias{governor}
\title{Did r correctly name governor}
\source{
General Social Survey https://gss.norc.org
}
\description{
governor
}
\details{
Question We would want to know how well known the different governmental leaders are around here? Could you tell me the name of the governor of this state?
}
\section{Values}{

\itemize{
\item \code{1} correct
\item \code{2} incorrect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1035/vshow}{GSS Data Explorer page for \code{governor}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}

Other Knowledge: 
\code{\link{schlhead}},
\code{\link{usrep}}

Other Political: 
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Knowledge}
\concept{Political}
\concept{Socio-political participation}
\keyword{variable}
