% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{govfnanc}
\alias{govfnanc}
\title{Govt should finance projects to create new jobs}
\source{
General Social Survey https://gss.norc.org
}
\description{
govfnanc
}
\details{
Question The government should finance projects to create new jobs, even if it might require a tax increase
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7552/vshow}{GSS Data Explorer page for \code{govfnanc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2245 \tab 180 \tab 566 \tab 540 \tab 312 \tab 100 \tab 64 \tab 3 \tab 22 \tab 4032 \cr
   Total \tab 2245 \tab 180 \tab 566 \tab 540 \tab 312 \tab 100 \tab 64 \tab 3 \tab 22 \tab 4032 \cr
}
}

\keyword{variable}
