% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{govngos}
\alias{govngos}
\title{Decision left to natl govt representative or citizen or}
\source{
General Social Survey https://gss.norc.org
}
\description{
govngos
}
\details{
Question Which of these two statements comes closer to your view?
}
\section{Values}{

\itemize{
\item \code{1} should be left to govt representatives
\item \code{2} citizen orgs should be involved
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4950/vshow}{GSS Data Explorer page for \code{govngos}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
