% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{govtbiz}
\alias{govtbiz}
\title{Government do more for environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
govtbiz
}
\details{
Question B. And which of these two groups do you think is making more effort to look after the environment:
}
\section{Values}{

\itemize{
\item \code{1} government
\item \code{2} or business and industry?
\item \code{3} both equally
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4785/vshow}{GSS Data Explorer page for \code{govtbiz}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}

Other Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
