% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_448.R
\name{govtinfo}
\alias{govtinfo}
\title{Should govt info be public}
\source{
General Social Survey https://gss.norc.org
}
\description{
govtinfo
}
\details{
Question (HAND CARD C6) 
Here is a scale from 0 to 10 where 0 is "All government information should be publically available, even if this means a risk to public security" and 10 is "public security should be given priority, even if this means limiting access to government information".
Where would you place yourself on such a scale?
ENTER THE RESPONDENT'S ANSWER, 0-10:
}
\section{Values}{

\itemize{
\item \code{0} 0 - all government information should be publically available, even if this meant a risk to public security
\item \code{10} 10 - public security should be given priority, even if this means limiting access to government information
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6317/vshow}{GSS Data Explorer page for \code{govtinfo}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cctv}},
\code{\link{cutdebt}},
\code{\link{emonitor}},
\code{\link{helpcrim}},
\code{\link{helpemp}},
\code{\link{helpenv}},
\code{\link{helphlth}},
\code{\link{helpold}},
\code{\link{helpsec}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
