% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{govtpow}
\alias{govtpow}
\title{Power of federal govt}
\source{
General Social Survey https://gss.norc.org
}
\description{
govtpow
}
\details{
Question And what about the federal government, does it have too much power or too little power?
}
\section{Values}{

\itemize{
\item \code{1} far too much power
\item \code{2} too much power
\item \code{3} about the right amount of power
\item \code{4} too little power
\item \code{5} far too little power
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1996 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4070/vshow}{GSS Data Explorer page for \code{govtpow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab far too much power \tab too much power \tab about the right amount of power \tab too little power \tab far too little power \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 104 \tab 261 \tab 240 \tab 25 \tab 2 \tab 35 \tab 10 \tab 1534 \cr
   1990 \tab 155 \tab 145 \tab 404 \tab 520 \tab 48 \tab 3 \tab 73 \tab 24 \tab 1372 \cr
   1996 \tab 1572 \tab 308 \tab 508 \tab 368 \tab 40 \tab 6 \tab 54 \tab 48 \tab 2904 \cr
   Total \tab 2584 \tab 557 \tab 1173 \tab 1128 \tab 113 \tab 11 \tab 162 \tab 82 \tab 5810 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
