% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{grace}
\alias{grace}
\title{Had powerful religious experience}
\source{
General Social Survey https://gss.norc.org
}
\description{
grace
}
\details{
Question How often have you had any of the following experiences?  READ EACH ITEM AND CIRCLE ONE CODE FOR EACH.
E. Felt as though you were very close to a powerful, spiritual force that seemed to lift you out of yourself.
}
\section{Values}{

\itemize{
\item \code{1} never in my life
\item \code{2} once or twice
\item \code{3} several times
\item \code{4} often
\item \code{8} cannot answer
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988, 1991 \cr
   Full \tab 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/320/vshow}{GSS Data Explorer page for \code{grace}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never in my life \tab once or twice \tab several times \tab often \tab can't choose \tab no answer \tab Total \cr
   1983 \tab - \tab 691 \tab 430 \tab 260 \tab 158 \tab 56 \tab 4 \tab 1599 \cr
   1984 \tab - \tab 857 \tab 294 \tab 197 \tab 94 \tab 12 \tab 19 \tab 1473 \cr
   1988 \tab - \tab 994 \tab 266 \tab 125 \tab 66 \tab 25 \tab 5 \tab 1481 \cr
   1989 \tab 531 \tab 695 \tab 174 \tab 74 \tab 45 \tab 15 \tab 3 \tab 1537 \cr
   1991 \tab 158 \tab 766 \tab 203 \tab 109 \tab 61 \tab 185 \tab 35 \tab 1517 \cr
   Total \tab 689 \tab 4003 \tab 1367 \tab 765 \tab 424 \tab 293 \tab 66 \tab 7607 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{spirits}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
