% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{gradtounder}
\alias{gradtounder}
\title{Ratio of graduate students to undergraduate students in college respondent atten}
\source{
General Social Survey https://gss.norc.org
}
\description{
gradtounder
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
D. Ratio of Graduate to Undergraduate Students
}
\section{Values}{

\itemize{
\item \verb{[1]} 0-.210 graduate student per undergraduate student
\item \verb{[2]} .211 or over
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/72/vshow}{GSS Data Explorer page for \code{gradtounder}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 0-.210 graduate student per undergraduate student \tab .211 or over \tab no answer \tab Total \cr
   2012 \tab 1494 \tab 161 \tab 316 \tab 3 \tab 1974 \cr
   2014 \tab 1898 \tab 230 \tab 406 \tab 4 \tab 2538 \cr
   2016 \tab 2109 \tab 248 \tab 503 \tab 7 \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 7849 \tab 639 \tab 1225 \tab 14 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
