% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{grassv}
\alias{grassv}
\title{Should marijuana be made legal (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
grassv
}
\details{
Question Do you think the use of marijuana should be made legal or not?
}
\section{Values}{

\itemize{
\item \code{1} should be legal
\item \code{2} should not be legal
\item \code{3} no opinion (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7595/vshow}{GSS Data Explorer page for \code{grassv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab should be legal \tab should not be legal \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2673 \tab 833 \tab 272 \tab 251 \tab 1 \tab 2 \tab 4032 \cr
   2022 \tab 2970 \tab 355 \tab 116 \tab 103 \tab - \tab - \tab 3544 \cr
   Total \tab 5643 \tab 1188 \tab 388 \tab 354 \tab 1 \tab 2 \tab 7576 \cr
}
}

\keyword{variable}
