% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{grassy}
\alias{grassy}
\title{Should marijuana be legal-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
grassy
}
\details{
Question Some people think the use of marijuana should be made legal.
Other people think marijuana use should not be made legal. Which do you favor?
}
\section{Values}{

\itemize{
\item \code{1} legal
\item \code{2} not legal
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/286/vshow}{GSS Data Explorer page for \code{grassy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab legal \tab not legal \tab don't know \tab no answer \tab Total \cr
   1978 \tab 761 \tab 242 \tab 501 \tab 27 \tab 1 \tab 1532 \cr
   Total \tab 761 \tab 242 \tab 501 \tab 27 \tab 1 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom}},
\code{\link{grass}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}},
\code{\link{wirtap}}
}
\concept{Core}
\keyword{variable}
