% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{grncon}
\alias{grncon}
\title{Concerned about environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
grncon
}
\details{
Question C. Generally speaking, how concerned are you about environmental issues?  Please tell me what you think, where 1 means you are not at all concerned and 5 means you are very concerned.
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all concerned
\item \code{5} 5 - very concerned
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4802/vshow}{GSS Data Explorer page for \code{grncon}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 - not at all concerned \tab 2 \tab 3 \tab 4 \tab 5 - very concerned \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 60 \tab 86 \tab 349 \tab 403 \tab 506 \tab 20 \tab 6 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 81 \tab 142 \tab 419 \tab 438 \tab 743 \tab 42 \tab 157 \tab 7 \tab 4032 \cr
   Total \tab 2617 \tab 141 \tab 228 \tab 768 \tab 841 \tab 1249 \tab 62 \tab 163 \tab 7 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
