% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{grncon}
\alias{grncon}
\title{Concerned about environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
grncon
}
\details{
Question C. Generally speaking, how concerned are you about environmental issues?  Please tell me what you think, where 1 means you are not at all concerned and 5 means you are very concerned.
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all concerned
\item \code{5} 5 - very concerned
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4802/vshow}{GSS Data Explorer page for \code{grncon}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}},
\code{\link{trust5}}

Other Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
