% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{grndemo}
\alias{grndemo}
\title{Protested for envir issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
grndemo
}
\details{
Question In the last five years, have you…
C. ...taken part in a protest or demonstration about an environmental issue?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, i have
\item \verb{[2]} no, i have not
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4776/vshow}{GSS Data Explorer page for \code{grndemo}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, i have \tab no, i have not \tab no answer \tab can't choose \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 40 \tab 1390 \tab 127 \tab - \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 39 \tab 1250 \tab 95 \tab 2 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 37 \tab 1123 \tab 115 \tab 1 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 31 \tab 1390 \tab 7 \tab 2 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 89 \tab 1728 \tab 196 \tab - \tab 16 \tab 4032 \cr
   Total \tab 5813 \tab 236 \tab 6881 \tab 540 \tab 5 \tab 16 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
