% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{grneffme}
\alias{grneffme}
\title{Environment effect everyday life}
\source{
General Social Survey https://gss.norc.org
}
\description{
grneffme
}
\details{
Question G2. (How much do you agree or disagree with each of these statements?) Environmental problems have a direct effect on my everyday life.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4809/vshow}{GSS Data Explorer page for \code{grneffme}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{peopgrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}
}
\concept{ISSP Environment}
\keyword{variable}
