% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{grnexagg}
\alias{grnexagg}
\title{Environmental threats exaggerated}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnexagg
}
\details{
Question How much do you agree or disagree with each of these statements?
C. Many of the claims about environmental threats are exaggerated.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4781/vshow}{GSS Data Explorer page for \code{grnexagg}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1541 \tab 70 \tab 254 \tab 333 \tab 365 \tab 128 \tab 53 \tab 73 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 85 \tab 393 \tab 257 \tab 488 \tab 120 \tab 79 \tab 8 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 106 \tab 282 \tab 354 \tab 604 \tab 431 \tab 64 \tab 180 \tab 8 \tab 4032 \cr
   Total \tab 4158 \tab 261 \tab 929 \tab 944 \tab 1457 \tab 679 \tab 196 \tab 261 \tab 8 \tab 8893 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{othssame}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
