% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{grngroup}
\alias{grngroup}
\title{Member of envir group}
\source{
General Social Survey https://gss.norc.org
}
\description{
grngroup
}
\details{
Question Are you a member of any group whose main aim is to preserve or protect the environment?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4773/vshow}{GSS Data Explorer page for \code{grngroup}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}},
\code{\link{usdoenuf}}

Other Environment: 
\code{\link{chemfree}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}},
\code{\link{usdoenuf}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
