% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{grnlaws}
\alias{grnlaws}
\title{Gov resp tomake industry less damaging}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnlaws
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to impose strict laws to make industry do less damage to the environment? PLEASE CHECK ONLY ONE BOX ONE EACH LINE.
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1996 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4094/vshow}{GSS Data Explorer page for \code{grnlaws}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 572 \tab 521 \tab 104 \tab 35 \tab 43 \tab 57 \tab 2904 \cr
   2006 \tab 2992 \tab 925 \tab 478 \tab 76 \tab 18 \tab 20 \tab 1 \tab 4510 \cr
   2016 \tab 1477 \tab 688 \tab 535 \tab 98 \tab 32 \tab 23 \tab 14 \tab 2867 \cr
   Total \tab 6041 \tab 2185 \tab 1534 \tab 278 \tab 85 \tab 86 \tab 72 \tab 10281 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
