% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{grnsign}
\alias{grnsign}
\title{Sign petition on envir issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnsign
}
\details{
Question In the last five years, have you…
A. ...signed a petition about an environmental issue?
}
\section{Values}{

\itemize{
\item \code{1} yes, i have
\item \code{2} no, i have not
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4774/vshow}{GSS Data Explorer page for \code{grnsign}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{nomeat}},
\code{\link{recycle}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
