% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{grnsol}
\alias{grnsol}
\title{Accept cut in living stnds to help envir?}
\source{
General Social Survey https://gss.norc.org
}
\description{
grnsol
}
\details{
Question C. And how willing would you be to accept cuts in your standard of living in order to protect the environment?
}
\section{Values}{

\itemize{
\item \code{1} very willing
\item \code{2} fairly willing
\item \code{3} neither willing nor unwilling
\item \code{4} fairly unwilling
\item \code{5} very unwilling
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4738/vshow}{GSS Data Explorer page for \code{grnsol}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very willing \tab fairly willing \tab neither willing nor unwilling \tab fairly unwilling \tab very unwilling \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 91 \tab 410 \tab 353 \tab 385 \tab 229 \tab 36 \tab 53 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 57 \tab 363 \tab 306 \tab 390 \tab 214 \tab 29 \tab 27 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 64 \tab 280 \tab 312 \tab 240 \tab 274 \tab 47 \tab 59 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 77 \tab 409 \tab 288 \tab 276 \tab 324 \tab 50 \tab 6 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 111 \tab 505 \tab 487 \tab 356 \tab 319 \tab 66 \tab 172 \tab 13 \tab 4032 \cr
   Total \tab 5813 \tab 400 \tab 1967 \tab 1746 \tab 1647 \tab 1360 \tab 228 \tab 317 \tab 13 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}},
\code{\link{toodifme}}
}
\concept{ISSP Environment}
\keyword{variable}
