% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_520.R
\name{grntaxes}
\alias{grntaxes}
\title{Pay higher taxes to help envir?}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntaxes
}
\details{
Question B. And how willing would you be to pay much higher taxes in order to protect the environment?
}
\section{Values}{

\itemize{
\item \code{1} very willing
\item \code{2} fairly willing
\item \code{3} neither willing nor unwilling
\item \code{4} fairly unwilling
\item \code{5} very unwilling
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4737/vshow}{GSS Data Explorer page for \code{grntaxes}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{resnatur}}

Other Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{resnatur}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
