% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{grntest5}
\alias{grntest5}
\title{All pesticides cause cancer}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntest5
}
\details{
Question And for each of these statements, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
E. All pesticides and chemicals used on food crops cause cancer in humans.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4750/vshow}{GSS Data Explorer page for \code{grntest5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 84 \tab 443 \tab 624 \tab 226 \tab 96 \tab 84 \tab 1606 \cr
   1994 \tab 1606 \tab 84 \tab 417 \tab 537 \tab 217 \tab 91 \tab 40 \tab 2992 \cr
   Total \tab 1655 \tab 168 \tab 860 \tab 1161 \tab 443 \tab 187 \tab 124 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
