% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{grntest6}
\alias{grntest6}
\title{Humans are main cause of plant & animal extinct}
\source{
General Social Survey https://gss.norc.org
}
\description{
grntest6
}
\details{
Question And for each of these statements, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
F. Human beings are the main cause of plant and animal species dying out.
}
\section{Values}{

\itemize{
\item \code{1} definitely true
\item \code{2} probably true
\item \code{3} probably not true
\item \code{4} definitely not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4751/vshow}{GSS Data Explorer page for \code{grntest6}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest7}},
\code{\link{nukefam}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
