% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{grpsport}
\alias{grpsport}
\title{Participated in sports group past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
grpsport
}
\details{
Question People sometimes belong to different kinds of groups or associations. The next few questions contain different types of groups. For each type of group, please select a response to say whether you have participated in the activities of this group in the past 12 months.
D. A sports group, hobby or leisure club
}
\section{Values}{

\itemize{
\item \code{1} i have participated more than twice
\item \code{2} i have participated once or twice
\item \code{3} i belong to such a group but never participate
\item \code{4} i do not belong to such a group
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4210/vshow}{GSS Data Explorer page for \code{grpsport}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpunion}},
\code{\link{neifrd}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
