% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{grpsprts}
\alias{grpsprts}
\title{R belongs to a sports, leisure, or cultural grp}
\source{
General Social Survey https://gss.norc.org
}
\description{
grpsprts
}
\details{
Question People sometimes belong to different kinds of groups or associations. For each type of group, please indicate whether you:
-Belong and actively participate,
-Belong but don’t actively participate,
-Used to belong but do not any more,
-Have never belonged to it.
D. A sports, leisure or cultural group
}
\section{Values}{

\itemize{
\item \code{1} belong and actively participate
\item \code{2} belong but dont participate
\item \code{3} used to belong
\item \code{4} never belonged
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4931/vshow}{GSS Data Explorer page for \code{grpsprts}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpwork}},
\code{\link{polopts}},
\code{\link{rghtsmin}},
\code{\link{solok}}
}
\concept{ISSP Citizenship}
\keyword{variable}
