% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{grtbooks}
\alias{grtbooks}
\title{Appeal of greatest books is univer'sal}
\source{
General Social Survey https://gss.norc.org
}
\description{
grtbooks
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
D. The greatest books are universal in their appeal:there is no "white literature," "black literature," or "Asian literature," there is only human literature.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1425/vshow}{GSS Data Explorer page for \code{grtbooks}}.
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{judgeart}},
\code{\link{modpaint}},
\code{\link{pclit}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}}
}
\concept{Culture}
\keyword{variable}
