% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_520.R
\name{grwthelp}
\alias{grwthelp}
\title{Amer needs econ grwth to protect envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
grwthelp
}
\details{
Question How much do you agree or disagree with each of the following statements.
C. In order to protect the environment, America needs economic growth.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4731/vshow}{GSS Data Explorer page for \code{grwthelp}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 78 \tab 661 \tab 335 \tab 339 \tab 35 \tab 48 \tab 61 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 74 \tab 512 \tab 337 \tab 339 \tab 39 \tab 51 \tab 34 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 95 \tab 451 \tab 313 \tab 245 \tab 35 \tab 60 \tab 77 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 127 \tab 682 \tab 256 \tab 259 \tab 29 \tab 71 \tab 6 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 143 \tab 637 \tab 617 \tab 297 \tab 60 \tab 89 \tab 171 \tab 15 \tab 4032 \cr
   Total \tab 5813 \tab 517 \tab 2943 \tab 1858 \tab 1479 \tab 198 \tab 319 \tab 349 \tab 15 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grntaxes}},
\code{\link{grwtharm}},
\code{\link{naturgod}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{resnatur}}
}
\concept{ISSP Environment}
\keyword{variable}
