% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_doc}
\alias{gss_doc}
\title{Data Dictionary for the GSS Cumulative Data File 1972-2024}
\format{
A tibble.

 * `variable` The name of the variable. (Character.)
 * `description` The variable's label. 'r' means 'Respondent'. (Character.)
 * `question` The question as asked. (Character)
 * `norc_id` The variable's ID on GSS Data Explorer. (Integer)
 * `norc_url` The URL for the variable's page on the GSS Data Explorer. (Character.)
 * `var_yrtab` Crosstab of responses by year. Stored in a list-column. The tibbles contain `year` values in the rows and value labels across their columns, with response freqencies in the cells as counts. (List.)
 * `yrballot_df` Crosstab of ballots by year. Stored in a list-column. (List)
 * `module_df` Tibble showing which module or modules the question belongs to. (List)
 * `subject_df` Tibble of NORC subject or topic classifications for the variable. (List)
 * `value_labels` Vector of values taken by the variable. (Character)
 * `var_type` The variable type of the variable in `gss_all`. (Character.)
 * `var_na_codes` The missing value codes associated with the variable in `gss_all`. (Character.)
}
\source{
National Opinion Research Center, [http://gss.norc.org](http://gss.norc.org).
}
\usage{
data(gss_doc)
}
\description{
A tibble containing information on the variables in the GSS Cumulative Data File.
}
\keyword{datasets}
