% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{gun}
\alias{gun}
\title{Ever threatened with gun or shot at}
\source{
General Social Survey https://gss.norc.org
}
\description{
gun
}
\details{
Question Have you ever been threatened with a gun, or shot at?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/661/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1256 \tab 2 \tab 246 \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 1229 \tab 3 \tab 257 \tab 1 \tab - \tab 1490 \cr
   1976 \tab - \tab 1241 \tab 1 \tab 256 \tab 1 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1220 \tab - \tab 311 \tab 1 \tab - \tab 1532 \cr
   1980 \tab - \tab 1156 \tab 3 \tab 308 \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1270 \tab 1 \tab 326 \tab 2 \tab - \tab 1599 \cr
   1984 \tab - \tab 1180 \tab 2 \tab 290 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1171 \tab - \tab 296 \tab 3 \tab - \tab 1470 \cr
   1987 \tab - \tab 1430 \tab 5 \tab 384 \tab - \tab - \tab 1819 \cr
   1988 \tab 484 \tab 779 \tab - \tab 217 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 832 \tab 3 \tab 200 \tab - \tab - \tab 1537 \cr
   1990 \tab 473 \tab 719 \tab 2 \tab 178 \tab - \tab - \tab 1372 \cr
   1991 \tab 500 \tab 810 \tab 6 \tab 201 \tab - \tab - \tab 1517 \cr
   1993 \tab 549 \tab 855 \tab 1 \tab 201 \tab - \tab - \tab 1606 \cr
   1994 \tab 2481 \tab 405 \tab 1 \tab 104 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 49476 \tab 15553 \tab 30 \tab 3775 \tab 12 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{antiwar}},
\code{\link{civright}},
\code{\link{gunage}},
\code{\link{hit}},
\code{\link{hitage}},
\code{\link{hitnum}},
\code{\link{prowar}},
\code{\link{school}},
\code{\link{strike}}

Other Guns: 
\code{\link{gunage}}
}
\concept{Core}
\concept{Guns}
\keyword{variable}
