% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{gunimp}
\alias{gunimp}
\title{Importance of gun control issue to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunimp
}
\details{
Question How important is the gun control issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \code{1} one of most imp
\item \code{2} important
\item \code{3} not very imp
\item \code{4} not important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1976, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/273/vshow}{GSS Data Explorer page for \code{gunimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not important \tab don't know \tab no answer \tab Total \cr
   1976 \tab - \tab 197 \tab 801 \tab 387 \tab 95 \tab 17 \tab 2 \tab 1499 \cr
   1984 \tab 989 \tab 81 \tab 252 \tab 121 \tab 20 \tab 6 \tab 4 \tab 1473 \cr
   Total \tab 989 \tab 278 \tab 1053 \tab 508 \tab 115 \tab 23 \tab 6 \tab 2972 \cr
}
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{guninfo}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
