% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_33.R
\name{guninfo}
\alias{guninfo}
\title{How much info does r have on gun control}
\source{
General Social Survey https://gss.norc.org
}
\description{
guninfo
}
\details{
Question How much information do you have about the gun control issue?
Do you have all the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \code{1} all info r needs
\item \code{2} most of the info
\item \code{3} some info
\item \code{4} very little info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/274/vshow}{GSS Data Explorer page for \code{guninfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all info r needs \tab most of the info \tab some info \tab very little info \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 67 \tab 105 \tab 194 \tab 110 \tab 4 \tab 4 \tab 1473 \cr
   Total \tab 989 \tab 67 \tab 105 \tab 194 \tab 110 \tab 4 \tab 4 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{capfirm}},
\code{\link{courts}},
\code{\link{courtsy}},
\code{\link{crimfirm}},
\code{\link{crimimp}},
\code{\link{criminfo}},
\code{\link{gunfirm}},
\code{\link{gunimp}},
\code{\link{gunlaw}}
}
\concept{Core}
\keyword{variable}
