% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_88.R
\name{gunner}
\alias{gunner}
\title{Women as air defense gunner in u.s.}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunner
}
\details{
Question I'm going to read you a list of jobs that people might have in the armed forces. Please tell me whether you think a woman should or should not be assigned to each job, assuming she is trained to do it.
H. Air defense missile gunner in the United States ...
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/796/vshow}{GSS Data Explorer page for \code{gunner}}.
}

\seealso{
Other Core: 
\code{\link{brass}},
\code{\link{fedirty}},
\code{\link{fefight}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{mechanic}},
\code{\link{nurse}},
\code{\link{transair}},
\code{\link{typist}}

Other Military: 
\code{\link{brass}},
\code{\link{fedirty}},
\code{\link{fefight}},
\code{\link{fightlnd}},
\code{\link{fightsea}},
\code{\link{mechanic}},
\code{\link{nurse}},
\code{\link{transair}},
\code{\link{typist}}
}
\concept{Core}
\concept{Military}
\keyword{variable}
