% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_75.R
\name{gunnum}
\alias{gunnum}
\title{If threatened or shot at--how many times}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunnum
}
\details{
Question Have you ever been threatened with a gun, or shot at?
B. How many times would you guess this has happened to you?
}
\section{Values}{

\itemize{
\item \code{1} once
\item \code{2} 2-3 times
\item \code{3} 4+ times
\item \code{8} not sure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab - \tab full \cr
   1975 \tab - \tab full \cr
   1976 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/663/vshow}{GSS Data Explorer page for \code{gunnum}}.
}

\seealso{
Other Core: 
\code{\link{hitbeatr}},
\code{\link{hitchild}},
\code{\link{hitdrunk}},
\code{\link{hitmarch}},
\code{\link{hitok}},
\code{\link{hitrobbr}},
\code{\link{polabuse}},
\code{\link{polhitok}},
\code{\link{polmurdr}}
}
\concept{Core}
\keyword{variable}
