% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{gunsales}
\alias{gunsales}
\title{Should background check be required for private gun sale}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunsales
}
\details{
Question In most states a gun owner may legally sell his or her gun without proof that the buyer has passed a criminal history check. How strongly do you favor or oppose a law that required private gun sales to be subject to the same background check requirements as sales by licensed dealers? Do you…
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3700/vshow}{GSS Data Explorer page for \code{gunsales}}.
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{rifles50}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\keyword{variable}
