% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{gunsdrug}
\alias{gunsdrug}
\title{Should penalty for illegal gun sale be tougher than drug}
\source{
General Social Survey https://gss.norc.org
}
\description{
gunsdrug
}
\details{
Question Should penalties for illegally selling guns be tougher, less tough, or about as tough as penalties for illegally selling drugs?
}
\section{Values}{

\itemize{
\item \code{1} tougher
\item \code{2} less tough
\item \code{3} about as tough
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3701/vshow}{GSS Data Explorer page for \code{gunsdrug}}.
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{othguns}},
\code{\link{rifles50}},
\code{\link{semiguns}}

Other Guns: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{othguns}},
\code{\link{rifles50}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\concept{Guns}
\keyword{variable}
