% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{gvinflu1}
\alias{gvinflu1}
\title{What peo or org have most inf}
\source{
General Social Survey https://gss.norc.org
}
\description{
gvinflu1
}
\details{
Question (HAND CARD C4)
Here is a list of people and organizations that can influence government actions.
Please read through this list and tell me the number corresponding to the one you think has the most influence on the actions of the American government  
:remarks: ''
}
\section{Values}{

\itemize{
\item \code{1} the media
\item \code{2} trade unions
\item \code{3} businesses, banks and industry
\item \code{4} religious organizations/authorities
\item \code{5} the military/army
\item \code{6} organized crime
\item \code{7} people who vote for the party/parties in government
\item \code{8} citizens in general
\item \code{9} civic and voluntary organizations
\item \code{10} international organizations (e.g. united nations, international monetary fund)
\item \code{11} can't choose
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6308/vshow}{GSS Data Explorer page for \code{gvinflu1}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{polgbeco}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
