% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{gvtapart}
\alias{gvtapart}
\title{Should govt policy maintain distinct cultures}
\source{
General Social Survey https://gss.norc.org
}
\description{
gvtapart
}
\details{
Question Some people say that it is better for America if different racial and ethnic groups maintain their distinct cultures. Others say that it is better if groups change so that they blend into the larger society as in the idea of a melting pot. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that racial and ethnic groups should maintain their distinct cultures, and a score of 7 as meaning that groups should change so that they blend into the larger society. What score between 1 and 7 comes closest to the way you feel? (CIRCLE ONE):
A. Should it be up to the government to help racial and ethnic groups maintain their distinct cultures, or should this be left up to the groups themselves?
}
\section{Values}{

\itemize{
\item \verb{[1]} govt to help
\item \verb{[2]} left up to groups
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1600/vshow}{GSS Data Explorer page for \code{gvtapart}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab govt to help \tab left up to groups \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2527 \tab 47 \tab 400 \tab 11 \tab 7 \tab 2992 \cr
   Total \tab 2527 \tab 47 \tab 400 \tab 11 \tab 7 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{gvtmelt}},
\code{\link{meltpot}}
}
\concept{Multiculturalism}
\keyword{variable}
