% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_371.R
\name{gwpol}
\alias{gwpol}
\title{Understanding of global warming by elected officials}
\source{
General Social Survey https://gss.norc.org
}
\description{
gwpol
}
\details{
Question The first issue is global warming. Global warming means a trend toward warmer temperatures throughoutthe world, with more extreme weather in many places and changes in food production that could affect our way of life. Some people believe that the burning of gasoline and other fossil fuels causes global warming. Others say that global warming has purely natural causes. Please look at Card B2. On a scale of 1 to 5, where 1 means "Very Well" and 5 means "Not at All", how well do the following groups understand the causes of global warming?
B. Elected Officials
}
\section{Values}{

\itemize{
\item \verb{[1]} very well
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3356/vshow}{GSS Data Explorer page for \code{gwpol}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very well \tab 2 \tab 3 \tab 4 \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3582 \tab 47 \tab 65 \tab 286 \tab 261 \tab 228 \tab 38 \tab 3 \tab 4510 \cr
   2010 \tab 1569 \tab 27 \tab 43 \tab 129 \tab 144 \tab 108 \tab 20 \tab 4 \tab 2044 \cr
   Total \tab 5151 \tab 74 \tab 108 \tab 415 \tab 405 \tab 336 \tab 58 \tab 7 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{gwbiz}},
\code{\link{gwsci}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsmath}},
\code{\link{hsphys}},
\code{\link{sciagrgw}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
