% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{h2olife}
\alias{h2olife}
\title{What property of water is most important}
\source{
General Social Survey https://gss.norc.org
}
\description{
h2olife
}
\details{
Question What property of water is most important for living organisms?
}
\section{Values}{

\itemize{
\item \code{1} it is odorless
\item \code{2} it does not conduct electricity
\item \code{3} it is tasteless
\item \code{4} it is liquid at most temperatures on earth
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3471/vshow}{GSS Data Explorer page for \code{h2olife}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}

Other Knowledge: 
\code{\link{anheat}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{stormtxt}}

Other Science: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
