% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{haggle}
\alias{haggle}
\title{Do you expect to negotiate for a lower price?}
\source{
General Social Survey https://gss.norc.org
}
\description{
haggle
}
\details{
Question When you go to such a sale, and when prices are posted, do you expect to pay the posted price, or do you expect to negotiate for a lower price?
}
\section{Values}{

\itemize{
\item \code{1} expect to pay posted price
\item \code{2} negotiate for lower price
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1875/vshow}{GSS Data Explorer page for \code{haggle}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab expect to pay posted price \tab negotiate for lower price \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1917 \tab 383 \tab 579 \tab 12 \tab 13 \tab 2904 \cr
   Total \tab 1917 \tab 383 \tab 579 \tab 12 \tab 13 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
