% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{hapboys}
\alias{hapboys}
\title{Boys' chances for a happy family life}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapboys
}
\details{
Question And what about boys growing up today? Do you think their chances for a happy family life will be better than yours, about the same, or worse than yours?
}
\section{Values}{

\itemize{
\item \code{1} better
\item \code{2} about the same
\item \code{3} worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1655/vshow}{GSS Data Explorer page for \code{hapboys}}.
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapgirls}},
\code{\link{rhmewrk}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{willwed1}},
\code{\link{willwed2}}

Other Family: 
\code{\link{hapgirls}}
}
\concept{Family}
\concept{Gender}
\keyword{variable}
