% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{hapcohab}
\alias{hapcohab}
\title{Happiness of relt with partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapcohab
}
\details{
Question B. Taking things all together, would you say that your relationship with your partner is very happy, pretty happy, or not too happy?
}
\section{Values}{

\itemize{
\item \code{1} very happy
\item \code{2} pretty happy
\item \code{3} not too happy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/436/vshow}{GSS Data Explorer page for \code{hapcohab}}.
}

\seealso{
Other Core: 
\code{\link{fair}},
\code{\link{health}},
\code{\link{helpful}},
\code{\link{life}},
\code{\link{satcity}},
\code{\link{sathobby}},
\code{\link{trust}},
\code{\link{trusty}}
}
\concept{Core}
\keyword{variable}
