% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{hapifwrk}
\alias{hapifwrk}
\title{Woman and family happier if she works}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapifwrk
}
\details{
Question Do you agree or disagree ...
D. A woman and her family will all be happier if she goes out to work.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4346/vshow}{GSS Data Explorer page for \code{hapifwrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 47 \tab 181 \tab 509 \tab 467 \tab 136 \tab 48 \tab 26 \tab 1481 \cr
   Total \tab 67 \tab 47 \tab 181 \tab 509 \tab 467 \tab 136 \tab 48 \tab 26 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famsuffr}},
\code{\link{fejobind}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{kidsuffr}},
\code{\link{mawrkwrm}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
