% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{happy7}
\alias{happy7}
\title{How happy r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
happy7
}
\details{
Question If you were to consider your life in general, how happy or unhappy would you say you are, on the whole?
}
\section{Values}{

\itemize{
\item \verb{[1]} completely happy
\item \verb{[2]} very happy
\item \verb{[3]} fairly happy
\item \verb{[4]} neither happy nor unhappy
\item \verb{[5]} fairly unhappy
\item \verb{[6]} very unhappy
\item \verb{[7]} completely unhappy
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4451/vshow}{GSS Data Explorer page for \code{happy7}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab completely happy \tab very happy \tab fairly happy \tab neither happy nor unhappy \tab fairly unhappy \tab very unhappy \tab completely unhappy \tab can't choose \tab no answer \tab skipped on web \tab not available in this year \tab Total \cr
   2002 \tab 1594 \tab 141 \tab 510 \tab 391 \tab 69 \tab 32 \tab 16 \tab 1 \tab 7 \tab 4 \tab - \tab - \tab 2765 \cr
   2012 \tab 672 \tab 148 \tab 563 \tab 440 \tab 77 \tab 35 \tab 16 \tab 5 \tab 11 \tab 7 \tab - \tab - \tab 1974 \cr
   2022 \tab 1220 \tab 282 \tab 782 \tab 865 \tab 175 \tab 123 \tab 34 \tab 8 \tab 43 \tab 5 \tab 7 \tab - \tab 3544 \cr
   Total \tab 3486 \tab 571 \tab 1855 \tab 1696 \tab 321 \tab 190 \tab 66 \tab 14 \tab 61 \tab 16 \tab 7 \tab 0 \tab 8283 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{Happiness}
\concept{ISSP Family & Gender Roles}
\concept{Split Ballots}
\keyword{variable}
