% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{happy7}
\alias{happy7}
\title{How happy r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
happy7
}
\details{
Question If you were to consider your life in general, how happy or unhappy would you say you are, on the whole?
}
\section{Values}{

\itemize{
\item \code{1} completely happy
\item \code{2} very happy
\item \code{3} fairly happy
\item \code{4} neither happy nor unhappy
\item \code{5} fairly unhappy
\item \code{6} very unhappy
\item \code{7} completely unhappy
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4451/vshow}{GSS Data Explorer page for \code{happy7}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab completely happy \tab very happy \tab fairly happy \tab neither happy nor unhappy \tab fairly unhappy \tab very unhappy \tab completely unhappy \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 1594 \tab 141 \tab 510 \tab 391 \tab 69 \tab 32 \tab 16 \tab 1 \tab 7 \tab 4 \tab - \tab 2765 \cr
   2012 \tab 672 \tab 148 \tab 563 \tab 440 \tab 77 \tab 35 \tab 16 \tab 5 \tab 11 \tab 7 \tab - \tab 1974 \cr
   2022 \tab 2386 \tab 120 \tab 425 \tab 413 \tab 94 \tab 50 \tab 17 \tab 2 \tab 29 \tab 5 \tab 3 \tab 3544 \cr
   Total \tab 4652 \tab 409 \tab 1498 \tab 1244 \tab 240 \tab 117 \tab 49 \tab 8 \tab 47 \tab 16 \tab 3 \tab 8283 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
