% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{happy7}
\alias{happy7}
\title{How happy r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
happy7
}
\details{
Question If you were to consider your life in general, how happy or unhappy would you say you are, on the whole?
}
\section{Values}{

\itemize{
\item \code{1} completely happy
\item \code{2} very happy
\item \code{3} fairly happy
\item \code{4} neither happy nor unhappy
\item \code{5} fairly unhappy
\item \code{6} very unhappy
\item \code{7} completely unhappy
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
   2022 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4451/vshow}{GSS Data Explorer page for \code{happy7}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{twoincs1}}
}
\concept{Happiness}
\concept{ISSP Family & Gender Roles}
\concept{Split Ballots}
\keyword{variable}
