% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{haprom}
\alias{haprom}
\title{Describe your romantic relationship?}
\source{
General Social Survey https://gss.norc.org
}
\description{
haprom
}
\details{
Question Taking things all together, how would you describe your romantic relationship? Would you say that your relationship is very happy, pretty happy, or not too happy?
}
\section{Values}{

\itemize{
\item \verb{[1]} very happy
\item \verb{[2]} pretty happy
\item \verb{[3]} not too happy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1651/vshow}{GSS Data Explorer page for \code{haprom}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very happy \tab pretty happy \tab not too happy \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2553 \tab 147 \tab 153 \tab 17 \tab 1 \tab 33 \tab 2904 \cr
   Total \tab 2553 \tab 147 \tab 153 \tab 17 \tab 1 \tab 33 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{livewith}},
\code{\link{romance}}
}
\concept{Gender}
\keyword{variable}
