% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{hapunhap}
\alias{hapunhap}
\title{Happy or unhappy with life today}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapunhap
}
\details{
Question If you were to consider your life in general these days, how happy or unhappy would you say you are, on the whole…
}
\section{Values}{

\itemize{
\item \verb{[1]} very happy
\item \verb{[2]} fairly happy
\item \verb{[3]} not very happy
\item \verb{[4]} not at all happy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998, 2002 \cr
   B/C/- \tab 2008 \cr
   C/-/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4627/vshow}{GSS Data Explorer page for \code{hapunhap}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab very happy \tab fairly happy \tab not very happy \tab not at all happy \tab don't know \tab no answer \tab 5 \tab 6 \tab 7 \tab skipped on web \tab Total \cr
   1991 \tab 158 \tab 501 \tab 737 \tab 96 \tab 12 \tab 9 \tab 4 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1998 \tab 1548 \tab 467 \tab 667 \tab 113 \tab 25 \tab 6 \tab 6 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2002 \tab 1616 \tab 464 \tab 575 \tab 76 \tab 14 \tab 11 \tab 9 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2008 \tab 658 \tab 587 \tab 677 \tab 83 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2018 \tab 1173 \tab 535 \tab 568 \tab 50 \tab 13 \tab 5 \tab 4 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 2378 \tab 162 \tab 357 \tab 452 \tab 81 \tab 14 \tab - \tab 73 \tab 17 \tab 6 \tab 4 \tab 3544 \cr
   Total \tab 7531 \tab 2716 \tab 3581 \tab 870 \tab 163 \tab 45 \tab 23 \tab 73 \tab 17 \tab 6 \tab 4 \tab 15029 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{deathpen}},
\code{\link{premars1}},
\code{\link{stiffpun}},
\code{\link{xmarsex1}}
}
\concept{ISSP Religion}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
