% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{hapunhap}
\alias{hapunhap}
\title{Happy or unhappy with life today}
\source{
General Social Survey https://gss.norc.org
}
\description{
hapunhap
}
\details{
Question If you were to consider your life in general these days, how happy or unhappy would you say you are, on the whole…
}
\section{Values}{

\itemize{
\item \code{1} very happy
\item \code{2} fairly happy
\item \code{3} not very happy
\item \code{4} not at all happy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2022 \tab C/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4627/vshow}{GSS Data Explorer page for \code{hapunhap}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{deathpen}},
\code{\link{premars1}},
\code{\link{stiffpun}},
\code{\link{xmarsex1}}
}
\concept{ISSP Religion}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
