% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{harmsgrn}
\alias{harmsgrn}
\title{Almost everything we do harms envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
harmsgrn
}
\details{
Question And please check one box for each of these statements to show how much you agree or disagree with it.
B. Almost everything we do in modern life harms the environment.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4726/vshow}{GSS Data Explorer page for \code{harmsgrn}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 95 \tab 669 \tab 244 \tab 431 \tab 45 \tab 19 \tab 54 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 72 \tab 553 \tab 240 \tab 403 \tab 46 \tab 29 \tab 43 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 78 \tab 457 \tab 249 \tab 339 \tab 48 \tab 26 \tab 79 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 96 \tab 601 \tab 258 \tab 385 \tab 41 \tab 44 \tab 5 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 135 \tab 639 \tab 520 \tab 451 \tab 53 \tab 50 \tab 168 \tab 13 \tab 4032 \cr
   Total \tab 5813 \tab 476 \tab 2919 \tab 1511 \tab 2009 \tab 233 \tab 168 \tab 349 \tab 13 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{anrights}},
\code{\link{grnecon}},
\code{\link{harmgood}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scifaith}},
\code{\link{scigrn}},
\code{\link{sciworse}}
}
\concept{ISSP Environment}
\keyword{variable}
