% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{harsexcl}
\alias{harsexcl}
\title{Since 18 r has ever found self the object of sexual advances from religious lead}
\source{
General Social Survey https://gss.norc.org
}
\description{
harsexcl
}
\details{
Question Now think about church, synagogue, or other religious settings. Since you turned 18, have you ever found yourself the object of sexual advances or propositions from a minister, priest, rabbi, or other clergyperson or religious leader who was not your spouse or significant other?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3936/vshow}{GSS Data Explorer page for \code{harsexcl}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 233 \tab 43 \tab 1715 \tab 3 \tab 29 \tab 2023 \cr
   Total \tab 233 \tab 43 \tab 1715 \tab 3 \tab 29 \tab 2023 \cr
}
}

\seealso{
Other Clergy/Sex: 
\code{\link{harsexjb}},
\code{\link{ownclerg}}
}
\concept{Clergy/Sex}
\keyword{variable}
