% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{harsexjb}
\alias{harsexjb}
\title{R has been the object of sexual advances/propositions from supervisor's}
\source{
General Social Survey https://gss.norc.org
}
\description{
harsexjb
}
\details{
Question Sometimes at work people find themselves the object of sexual advances or propositions from supervisors. The advances sometimes involve physical contact and sometimes just involve sexual conversations. Have you experienced such advances?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3935/vshow}{GSS Data Explorer page for \code{harsexjb}}.
}

\seealso{
Other Clergy/Sex: 
\code{\link{harsexcl}},
\code{\link{ownclerg}}
}
\concept{Clergy/Sex}
\keyword{variable}
