% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{health1}
\alias{health1}
\title{R's health in general}
\source{
General Social Survey https://gss.norc.org
}
\description{
health1
}
\details{
Question Would you say that in general your health is Excellent, Very good, Good, Fair, or Poor?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/C/- \tab partial \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2826/vshow}{GSS Data Explorer page for \code{health1}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{backpain}},
\code{\link{chngtme}},
\code{\link{condemnd}},
\code{\link{cowrkhlp}},
\code{\link{cowrkint}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{fairearn}},
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{fringeok}},
\code{\link{handmove}},
\code{\link{haveinfo}},
\code{\link{height}},
\code{\link{hlpequip}},
\code{\link{hlthdays}},
\code{\link{hrsrelax}},
\code{\link{hurtatwk}},
\code{\link{hvylift}},
\code{\link{hyperten}},
\code{\link{jobfind1}},
\code{\link{jobsecok}},
\code{\link{knowschd}},
\code{\link{knowschd1}},
\code{\link{knowwhat}},
\code{\link{laidoff}},
\code{\link{learnnew}},
\code{\link{lotofsay}},
\code{\link{manvsemp}},
\code{\link{misswork}},
\code{\link{mntlhlth}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{myskills}},
\code{\link{ntwkhard}},
\code{\link{opdevel}},
\code{\link{overwork}},
\code{\link{painarms}},
\code{\link{partteam}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{prodctiv}},
\code{\link{promtefr}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{respect}},
\code{\link{rincblls}},
\code{\link{safefrst}},
\code{\link{safehlth}},
\code{\link{safetywk}},
\code{\link{satjob1}},
\code{\link{secondwk}},
\code{\link{setthngs}},
\code{\link{slpprblm}},
\code{\link{spvtrfair}},
\code{\link{stress12}},
\code{\link{strredpg}},
\code{\link{supcares}},
\code{\link{suphelp}},
\code{\link{teamsafe}},
\code{\link{toofewwk}},
\code{\link{trainops}},
\code{\link{trdunion}},
\code{\link{trustman}},
\code{\link{trynewjb}},
\code{\link{usedup}},
\code{\link{usetech}},
\code{\link{waypaid}},
\code{\link{weight}},
\code{\link{whywkhme}},
\code{\link{wkageism}},
\code{\link{wkbonus}},
\code{\link{wkdecide}},
\code{\link{wkfreedm}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkpraise}},
\code{\link{wkracism}},
\code{\link{wksexism}},
\code{\link{wksmooth}},
\code{\link{wktopsat}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}},
\code{\link{wrksched}},
\code{\link{wrktime}},
\code{\link{wrktype}},
\code{\link{yearsjob}}

Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{denyrais}},
\code{\link{difstand}},
\code{\link{eharaswk}},
\code{\link{gdjobsec}},
\code{\link{hotargus}},
\code{\link{ignorwk}},
\code{\link{jobsecok}},
\code{\link{jokeswk}},
\code{\link{lackinfo}},
\code{\link{liedcwkr}},
\code{\link{lookaway}},
\code{\link{manvsemp}},
\code{\link{mntlhlth}},
\code{\link{numemps}},
\code{\link{numemps_0422}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{physharm}},
\code{\link{putdown}},
\code{\link{reliedon}},
\code{\link{rptprobs}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{supcares}},
\code{\link{treatres}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkracism}},
\code{\link{wkrspns}},
\code{\link{wkvsfam}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
