% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{hefinfo1}
\alias{hefinfo1}
\title{Number of household roster informant}
\source{
General Social Survey https://gss.norc.org
}
\description{
hefinfo1
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} 1st person
\item \code{2} 2nd person
\item \code{3} 3rd person
\item \code{4} 4th person
\item \code{5} 5th person
\item \code{6} 6th person
\item \code{7} 7th person
\item \code{8} 8th person
\item \code{9} 9th person
\item \code{10} 10th person
\item \code{11} 11th person
\item \code{12} 12th person
\item \code{13} 13th person
\item \code{14} 14th person
\item \code{22} non-household person
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
