% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{helpblk}
\alias{helpblk}
\title{Should govt aid black people?}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpblk
}
\details{
Question Now look at CARD BF. Some people think that Black people have been discriminated against for so long that the government has a special obligation to help improve their living standards. Others believe that the government should not be giving special treatment to Black people.A. Where would you place yourself on this scale, or haven't you made up your mind on this?
}
\section{Values}{

\itemize{
\item \code{1} government should help
\item \code{3} agree with both
\item \code{5} no special treatment
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab B/C/- \tab partial \cr
   1989 \tab B/C/- \tab partial \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab B/C/- \tab partial \cr
   1993 \tab B/C/- \tab partial \cr
   1994 \tab B/C/- \tab partial \cr
   1996 \tab B/C/- \tab partial \cr
   1998 \tab B/C/- \tab partial \cr
   2000 \tab B/C/- \tab partial \cr
   2002 \tab B/C/- \tab partial \cr
   2004 \tab B/C/- \tab partial \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab B/C/- \tab partial \cr
   2010 \tab B/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/847/vshow}{GSS Data Explorer page for \code{helpblk}}.
}

\seealso{
Other Core: 
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{obmepax}},
\code{\link{obmewar}}
}
\concept{Core}
\keyword{variable}
