% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{helpfrds}
\alias{helpfrds}
\title{Should better off ppl help less well-off friend}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpfrds
}
\details{
Question Please select a response to indicate how much you agree or disagree with each of the following statements.
C. People who are better off should help friends who are less well off
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4231/vshow}{GSS Data Explorer page for \code{helpfrds}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{demands}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{usefrds}},
\code{\link{wantbest}}

Other Altruism: 
\code{\link{usefrds}}
}
\concept{Altruism}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
