% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{helpfulnv}
\alias{helpfulnv}
\title{People helpful or looking out for selves (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpfulnv
}
\details{
Question Would you say that most of the time people try to be helpful, or that they are mostly just looking out for themselves?
}
\section{Values}{

\itemize{
\item \code{1} try to be helpful
\item \code{2} looking out for themselves
\item \code{3} depends (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7611/vshow}{GSS Data Explorer page for \code{helpfulnv}}.
}

\keyword{variable}
