% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_143.R
\name{helporg}
\alias{helporg}
\title{Willing work harder than have to help org.}
\source{
General Social Survey https://gss.norc.org
}
\description{
helporg
}
\details{
Question Please tell me how much you agree or disagree with the following statements. Would you say that you strongly agree, agree, disagree, or strongly disagree?
B. I am willing to work harder than I have to in order to help this organization succeed.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1341/vshow}{GSS Data Explorer page for \code{helporg}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 339 \tab 466 \tab 83 \tab 7 \tab 5 \tab 12 \tab 1517 \cr
   Total \tab 605 \tab 339 \tab 466 \tab 83 \tab 7 \tab 5 \tab 12 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{chngeorg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{samevals}},
\code{\link{stayorg1}},
\code{\link{stayorg2}},
\code{\link{wrkmuch}}
}
\concept{Work Organizations}
\keyword{variable}
