% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{helporg1}
\alias{helporg1}
\title{Willing to work harder for the firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
helporg1
}
\details{
Question To what extent do you agree or disagree with each of the following statements?
A. I am willing to work harder than I have to in order to help the firm or organization I work for succeed.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4585/vshow}{GSS Data Explorer page for \code{helporg1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 190 \tab 434 \tab 141 \tab 43 \tab 4 \tab 9 \tab 44 \tab 2832 \cr
   2006 \tab 3492 \tab 364 \tab 447 \tab 138 \tab 45 \tab 17 \tab 4 \tab 3 \tab 4510 \cr
   2016 \tab 1927 \tab 338 \tab 413 \tab 132 \tab 36 \tab 10 \tab 4 \tab 7 \tab 2867 \cr
   Total \tab 7386 \tab 892 \tab 1294 \tab 411 \tab 124 \tab 31 \tab 17 \tab 54 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
