% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{helpsick}
\alias{helpsick}
\title{Should govt help pay for medical care?}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpsick
}
\details{
Question Look at CARD BE. In general, some people think that it is the responsibility of the government in Washington to see to it that people have help in paying for doctors and hospital bills. Others think that these matters are not the responsibility of the federal government and that people should take care of these things themselves.
A. Where would you place yourself on this scale, or haven't you made up your mind on this?
}
\section{Values}{

\itemize{
\item \code{1} government should help
\item \code{3} agree with both
\item \code{5} people should care for themselves
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1975, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/846/vshow}{GSS Data Explorer page for \code{helpsick}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab government should help \tab 2 \tab agree with both \tab 4 \tab people should care for themselves \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1975 \tab - \tab 532 \tab 188 \tab 423 \tab 115 \tab 195 \tab 31 \tab 6 \tab - \tab 1490 \cr
   1983 \tab - \tab 413 \tab 300 \tab 504 \tab 162 \tab 161 \tab 53 \tab 6 \tab - \tab 1599 \cr
   1984 \tab - \tab 347 \tab 269 \tab 506 \tab 176 \tab 111 \tab 40 \tab 24 \tab - \tab 1473 \cr
   1986 \tab - \tab 416 \tab 292 \tab 462 \tab 164 \tab 93 \tab 36 \tab 7 \tab - \tab 1470 \cr
   1987 \tab - \tab 549 \tab 353 \tab 610 \tab 141 \tab 121 \tab 39 \tab 6 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 264 \tab 214 \tab 345 \tab 85 \tab 64 \tab 21 \tab 4 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 314 \tab 231 \tab 310 \tab 78 \tab 77 \tab 21 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 266 \tab 229 \tab 270 \tab 67 \tab 39 \tab 26 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 321 \tab 250 \tab 271 \tab 93 \tab 58 \tab 20 \tab 4 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 297 \tab 234 \tab 332 \tab 98 \tab 65 \tab 23 \tab 8 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 503 \tab 420 \tab 620 \tab 217 \tab 185 \tab 53 \tab 13 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 518 \tab 403 \tab 622 \tab 198 \tab 126 \tab 51 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 473 \tab 423 \tab 610 \tab 185 \tab 158 \tab 56 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 539 \tab 402 \tab 586 \tab 194 \tab 109 \tab 57 \tab 9 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 287 \tab 175 \tab 308 \tab 65 \tab 63 \tab 13 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 280 \tab 175 \tab 274 \tab 66 \tab 69 \tab 11 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 665 \tab 367 \tab 634 \tab 166 \tab 132 \tab 25 \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 471 \tab 252 \tab 384 \tab 125 \tab 99 \tab 32 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 402 \tab 224 \tab 422 \tab 142 \tab 136 \tab 45 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 384 \tab 228 \tab 410 \tab 153 \tab 122 \tab 39 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 412 \tab 311 \tab 533 \tab 175 \tab 209 \tab 46 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 913 \tab 587 \tab 364 \tab 599 \tab 190 \tab 156 \tab 55 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 528 \tab 306 \tab 448 \tab 129 \tab 117 \tab 33 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 1377 \tab 917 \tab 471 \tab 793 \tab 227 \tab 219 \tab 4 \tab 1 \tab 23 \tab 4032 \cr
   2022 \tab 1174 \tab 949 \tab 381 \tab 687 \tab 140 \tab 166 \tab 33 \tab 4 \tab 10 \tab 3544 \cr
   Total \tab 19681 \tab 11634 \tab 7462 \tab 11963 \tab 3551 \tab 3050 \tab 863 \tab 129 \tab 33 \tab 58366 \cr
}
}

\seealso{
Other Core: 
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{obmepax}},
\code{\link{obmewar}}
}
\concept{Core}
\keyword{variable}
