% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{helpwrld}
\alias{helpwrld}
\title{How important to help worse off ppl in rest of world}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpwrld
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
I. To help people in the rest of the world who are worse off than yourself
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4915/vshow}{GSS Data Explorer page for \code{helpwrld}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 69 \tab 92 \tab 148 \tab 274 \tab 335 \tab 216 \tab 326 \tab 8 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 81 \tab 93 \tab 124 \tab 244 \tab 262 \tab 154 \tab 271 \tab 34 \tab 1 \tab 2538 \cr
   Total \tab 2614 \tab 150 \tab 185 \tab 272 \tab 518 \tab 597 \tab 370 \tab 597 \tab 42 \tab 5 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{relmeet}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
