% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{hguncrim}
\alias{hguncrim}
\title{More restrictions on handgun decrease crime}
\source{
General Social Survey https://gss.norc.org
}
\description{
hguncrim
}
\details{
Question Some people argue that more restrictions on handguns would decrease violent crime by making it harder for criminals to get handguns. Other people argue that more restrictions on handguns would increase violent crime by making it harder for law-abiding citizens to defend themselves with handguns. Which of the following statements is closer to your own opinion?
}
\section{Values}{

\itemize{
\item \code{1} more restriction would decrease crime
\item \code{2} more restriction would increase crime
\item \code{3} unsure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3058/vshow}{GSS Data Explorer page for \code{hguncrim}}.
}

\seealso{
Other Guns: 
\code{\link{crimdown}},
\code{\link{crimup}},
\code{\link{hgunlaw}}
}
\concept{Guns}
\keyword{variable}
