% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_341.R
\name{hguncrim}
\alias{hguncrim}
\title{More restrictions on handgun decrease crime}
\source{
General Social Survey https://gss.norc.org
}
\description{
hguncrim
}
\details{
Question Some people argue that more restrictions on handguns would decrease violent crime by making it harder for criminals to get handguns. Other people argue that more restrictions on handguns would increase violent crime by making it harder for law-abiding citizens to defend themselves with handguns. Which of the following statements is closer to your own opinion?
}
\section{Values}{

\itemize{
\item \verb{[1]} more restriction would decrease crime
\item \verb{[2]} more restriction would increase crime
\item \verb{[3]} unsure
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3058/vshow}{GSS Data Explorer page for \code{hguncrim}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab more restriction would decrease crime \tab more restriction would increase crime \tab unsure \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1499 \tab 589 \tab 251 \tab 455 \tab 10 \tab 8 \tab 2812 \cr
   Total \tab 1499 \tab 589 \tab 251 \tab 455 \tab 10 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Guns: 
\code{\link{crimdown}},
\code{\link{crimup}},
\code{\link{hgunlaw}}
}
\concept{Guns}
\keyword{variable}
