% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{hhrace}
\alias{hhrace}
\title{Race of household}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhrace
}
\details{
Question Race of household.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black
\item \code{3} american indian
\item \code{4} asiatic, oriental
\item \code{5} other, mixed, not able to observe
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5431/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab asiatic, oriental \tab black \tab white \tab american indian \tab other, mixed, not able to observe \tab no answer \tab don't know \tab not available in this year \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 755 \tab 3 \tab 66 \tab 666 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 755 \tab 4 \tab 69 \tab 670 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 14 \tab 176 \tab 1337 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 15 \tab 160 \tab 1351 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 8 \tab 138 \tab 1294 \tab 4 \tab 6 \tab 18 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 12 \tab 491 \tab 1336 \tab 2 \tab 13 \tab 6 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 12 \tab 161 \tab 1383 \tab 1 \tab 40 \tab 2 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 13 \tab 166 \tab 1236 \tab 14 \tab 28 \tab 16 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 14 \tab 148 \tab 1328 \tab 9 \tab 24 \tab 11 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 10 \tab 180 \tab 1246 \tab 12 \tab 20 \tab 2 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 11 \tab 518 \tab 1215 \tab 16 \tab 40 \tab 19 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 15 \tab 185 \tab 1237 \tab 15 \tab 23 \tab 6 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 18 \tab 156 \tab 1302 \tab 10 \tab 32 \tab 19 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 10 \tab 157 \tab 1127 \tab 21 \tab 38 \tab 18 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 14 \tab 194 \tab 1210 \tab 10 \tab 34 \tab 55 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 28 \tab 172 \tab 1310 \tab 15 \tab 37 \tab 44 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 40 \tab 366 \tab 2392 \tab 11 \tab 104 \tab 79 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 41 \tab 366 \tab 2219 \tab 15 \tab 115 \tab 147 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 50 \tab 354 \tab 2179 \tab 12 \tab 104 \tab 133 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 51 \tab 397 \tab 2093 \tab 12 \tab 88 \tab 176 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 44 \tab 384 \tab 2173 \tab 8 \tab 156 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 64 \tab 361 \tab 2217 \tab 4 \tab 153 \tab - \tab 13 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 97 \tab 612 \tab 3501 \tab 14 \tab 279 \tab 1 \tab 6 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 38 \tab 273 \tab 1587 \tab 8 \tab 107 \tab 3 \tab 7 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 46 \tab 290 \tab 1549 \tab 2 \tab 150 \tab 3 \tab 4 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 43 \tab 288 \tab 1443 \tab 20 \tab 175 \tab - \tab 5 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 46 \tab 373 \tab 1797 \tab 21 \tab 294 \tab - \tab 7 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 61 \tab 461 \tab 2085 \tab 22 \tab 226 \tab - \tab 12 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 49 \tab 386 \tab 1636 \tab 11 \tab 252 \tab 3 \tab 11 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab 1589 \tab 42 \tab 323 \tab 1295 \tab 13 \tab 214 \tab 2 \tab 66 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 7700 \tab 913 \tab 8371 \tab 47414 \tab 304 \tab 2760 \tab 763 \tab 133 \tab 4032 \tab 3309 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{hefinfo}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{respnum}}

Other Household: 
\code{\link{hefinfo}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{respnum}}
}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
