% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{hhtype}
\alias{hhtype}
\title{Household type}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype
}
\details{
Question Household Type.
}
\section{Values}{

\itemize{
\item \code{1} 1adlt,0kids
\item \code{2} 1adlt,1+kids
\item \code{3} 2adlts,mar,0kids
\item \code{4} 2adlts,mar,1+kids
\item \code{5} 2adlts,dkmar,0kids
\item \code{6} 2adlts,dkmar,1+kids
\item \code{7} 2as,ntmar,ntrel,opsex,0ks
\item \code{8} 2as,ntmar,ntrel,opsex,1+ks
\item \code{9} 2as,ntmar,ntrel,smsex,0ks
\item \code{10} 2as,ntmar,ntrel,smsex,1+ks
\item \code{11} 2adlts,ntmar,rel,0kids
\item \code{12} 2adlts,ntmar,rel,1+kids
\item \code{13} 2adlts,ntmar,dkrel,0kids
\item \code{14} 2adlts,ntmar,dkrel,1+kids
\item \code{15} 2as,ntmar,ntrel,dksex,0ks
\item \code{16} 2as,ntmar,ntrel,dksex,1+ks
\item \code{20} 3adlt,0mar,sp,0kids
\item \code{21} 3adlts,0mar,sp,rel,0kids
\item \code{22} 3adlts,0mar,rel,sp,0kids
\item \code{23} 3adlts,0mar,rel,0kids
\item \code{24} 3adlts,0mar,sp,ntrel,0kids
\item \code{25} 3adlts,0mar,ntrel,sp,0kids
\item \code{26} 3adlts,0mar,rel,0kids
\item \code{27} 3adlts,0mar,ntrel,rel,0kids
\item \code{28} 3adlts,0mar,ntrel,0kids
\item \code{30} 3adlts,1mar,sp,0kids
\item \code{31} 3adlts,1mar,sp,rel,0kids
\item \code{32} 3adlts,1mar,rel,sp,0kids
\item \code{33} 3adlts,1mar,rel,rel,0kids
\item \code{34} 3adlts,1mar,sp,ntrel,0kids
\item \code{35} 3adlts,1mar,ntrel,sp,0kids
\item \code{36} 3adlts,1mar,rel,ntrel,0kids
\item \code{37} 3adlts,1mar,ntrel,rel,0kids
\item \code{38} 3adlts,1mar,ntrel,0kids
\item \code{40} 3adlts,2mar,sp,0kids
\item \code{41} 3adlts,2mar,sp,rel,0kids
\item \code{42} 3adlts,2mar,rel,sp,0kids
\item \code{43} 3adlts,2mar,rel,0kids
\item \code{44} 3adlts,2mar,sp,ntrel,0kids
\item \code{45} 3adlts,2mar,rel,ntrel,0kids
\item \code{46} 3adlts,2mar,rel,ntrel,0kids
\item \code{47} 3adlts,2mar,ntrel,rel,0kids
\item \code{48} 3adlts,2mar,ntrel,0kids
\item \code{50} 3adlts,3mar,sp,0kids
\item \code{51} 3adlts,3mar,sp,rel,0kids
\item \code{52} 3adlts,3mar,sp,rel,0kids
\item \code{53} 3adlts,3mar,rel,0kids
\item \code{54} 3adlts,3mar,sp,ntrel,0kids
\item \code{55} 3adlts,3mar,ntrel,sp,0kids
\item \code{56} 3adlts,3mar,rel,ntrel,0kids
\item \code{57} 3adlts,3mar,ntrel,rel,0kids
\item \code{58} 3adlts,3mar,ntrel,0kids
\item \code{100} 4+adlts,0mar,0kids
\item \code{101} 4+adlts,1mar,0kids
\item \code{102} 4+adlts,2mar,0kids
\item \code{103} 4+adlts,3mar,0kids
\item \code{104} 4+adlts,4+mar,0kids
\item \code{120} 3adlt,0mar,sp,1+kids
\item \code{121} 3adlts,0mar,sp,rel,1+kids
\item \code{122} 3adlts,0mar,rel,sp,1+kids
\item \code{123} 3adlts,0mar,rel,1+kids
\item \code{124} 3adlts,0mar,sp,ntrel,1+kids
\item \code{125} 3adlts,0mar,ntrel,sp,1+kids
\item \code{126} 3adlts,0mar,rel,1+kids
\item \code{127} 3adlts,0mar,ntrel,rel,1+kids
\item \code{128} 3adlts,0mar,ntrel,1+kids
\item \code{130} 3adlts,1mar,sp,1+kids
\item \code{131} 3adlts,1mar,sp,rel,1+kids
\item \code{132} 3adlts,1mar,rel,sp,1+kids
\item \code{133} 3adlts,1mar,rel,rel,1+kids
\item \code{134} 3adlts,1mar,sp,ntrel,1+kids
\item \code{135} 3adlts,1mar,ntrel,sp,1+kids
\item \code{136} 3adlts,1mar,rel,ntrel,1+kids
\item \code{137} 3adlts,1mar,ntrel,rel,1+kids
\item \code{138} 3adlts,1mar,ntrel,1+kids
\item \code{140} 3adlts,2mar,sp,1+kids
\item \code{141} 3adlts,2mar,sp,rel,1+kids
\item \code{142} 3adlts,2mar,rel,sp,1+kids
\item \code{143} 3adlts,2mar,rel,1+kids
\item \code{144} 3adlts,2mar,sp,ntrel,1+kids
\item \code{145} 3adlts,2mar,rel,ntrel,1+kids
\item \code{146} 3adlts,2mar,rel,ntrel,1+kids
\item \code{147} 3adlts,2mar,ntrel,rel,1+kids
\item \code{148} 3adlts,2mar,ntrel,1+kids
\item \code{150} 3adlts,3mar,sp,1+kids
\item \code{151} 3adlts,3mar,sp,rel,1+kids
\item \code{152} 3adlts,3mar,sp,rel,1+kids
\item \code{153} 3adlts,3mar,rel,1+kids
\item \code{154} 3adlts,3mar,sp,ntrel,1+kids
\item \code{155} 3adlts,3mar,ntrel,sp,1+kids
\item \code{156} 3adlts,3mar,rel,ntrel,1+kids
\item \code{157} 3adlts,3mar,ntrel,rel,1+kids
\item \code{158} 3adlts,3mar,ntrel,1+kids
\item \code{200} 4+adlts,0mar,1+kids
\item \code{201} 4+adlts,1mar,1+kids
\item \code{202} 4+adlts,2mar,1+kids
\item \code{203} 4+adlts,3mar,1+kids
\item \code{204} 4+adlts,4+mar,1+kids
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away1}},
\code{\link{away10}},
\code{\link{away11}},
\code{\link{away12}},
\code{\link{away13}},
\code{\link{away14}},
\code{\link{away2}},
\code{\link{away3}},
\code{\link{away4}},
\code{\link{away5}},
\code{\link{away6}},
\code{\link{away7}},
\code{\link{away8}},
\code{\link{away9}},
\code{\link{famgen}},
\code{\link{famgen_exp}},
\code{\link{gender1}},
\code{\link{gender10}},
\code{\link{gender11}},
\code{\link{gender12}},
\code{\link{gender13}},
\code{\link{gender14}},
\code{\link{gender2}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{gender5}},
\code{\link{gender6}},
\code{\link{gender7}},
\code{\link{gender8}},
\code{\link{gender9}},
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{hhtype1}},
\code{\link{mar1}},
\code{\link{mar10}},
\code{\link{mar11}},
\code{\link{mar12}},
\code{\link{mar13}},
\code{\link{mar14}},
\code{\link{mar2}},
\code{\link{mar3}},
\code{\link{mar4}},
\code{\link{mar5}},
\code{\link{mar6}},
\code{\link{mar7}},
\code{\link{mar8}},
\code{\link{mar9}},
\code{\link{old1}},
\code{\link{old10}},
\code{\link{old11}},
\code{\link{old12}},
\code{\link{old13}},
\code{\link{old14}},
\code{\link{old2}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{old5}},
\code{\link{old6}},
\code{\link{old7}},
\code{\link{old8}},
\code{\link{old9}},
\code{\link{relate1}},
\code{\link{relate10}},
\code{\link{relate11}},
\code{\link{relate12}},
\code{\link{relate13}},
\code{\link{relate14}},
\code{\link{relate2}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{relate5}},
\code{\link{relate6}},
\code{\link{relate7}},
\code{\link{relate8}},
\code{\link{relate9}},
\code{\link{relhh1}},
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh13}},
\code{\link{relhh14}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{relhh5}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}},
\code{\link{relhhd1}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{relhhd8}},
\code{\link{relhhd9}},
\code{\link{relsp1}},
\code{\link{relsp10}},
\code{\link{relsp11}},
\code{\link{relsp12}},
\code{\link{relsp13}},
\code{\link{relsp14}},
\code{\link{relsp2}},
\code{\link{relsp3}},
\code{\link{relsp4}},
\code{\link{relsp5}},
\code{\link{relsp6}},
\code{\link{relsp7}},
\code{\link{relsp8}},
\code{\link{relsp9}},
\code{\link{respnum}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}},
\code{\link{where1}},
\code{\link{where10}},
\code{\link{where11}},
\code{\link{where12}},
\code{\link{where13}},
\code{\link{where14}},
\code{\link{where2}},
\code{\link{where3}},
\code{\link{where4}},
\code{\link{where5}},
\code{\link{where6}},
\code{\link{where7}},
\code{\link{where8}},
\code{\link{where9}}
}
\concept{Household Members and Structure}
\keyword{variable}
