% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{hhtype}
\alias{hhtype}
\title{Household type}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype
}
\details{
Question Household Type.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1adlt,0kids
\item \verb{[2]} 1adlt,1+kids
\item \verb{[3]} 2adlts,mar,0kids
\item \verb{[4]} 2adlts,mar,1+kids
\item \verb{[5]} 2adlts,dkmar,0kids
\item \verb{[6]} 2adlts,dkmar,1+kids
\item \verb{[7]} 2as,ntmar,ntrel,opsex,0ks
\item \verb{[8]} 2as,ntmar,ntrel,opsex,1+ks
\item \verb{[9]} 2as,ntmar,ntrel,smsex,0ks
\item \verb{[10]} 2as,ntmar,ntrel,smsex,1+ks
\item \verb{[11]} 2adlts,ntmar,rel,0kids
\item \verb{[12]} 2adlts,ntmar,rel,1+kids
\item \verb{[13]} 2adlts,ntmar,dkrel,0kids
\item \verb{[14]} 2adlts,ntmar,dkrel,1+kids
\item \verb{[15]} 2as,ntmar,ntrel,dksex,0ks
\item \verb{[16]} 2as,ntmar,ntrel,dksex,1+ks
\item \verb{[20]} 3adlt,0mar,sp,0kids
\item \verb{[21]} 3adlts,0mar,sp,rel,0kids
\item \verb{[22]} 3adlts,0mar,rel,sp,0kids
\item \verb{[23]} 3adlts,0mar,rel,0kids
\item \verb{[24]} 3adlts,0mar,sp,ntrel,0kids
\item \verb{[25]} 3adlts,0mar,ntrel,sp,0kids
\item \verb{[26]} 3adlts,0mar,rel,0kids
\item \verb{[27]} 3adlts,0mar,ntrel,rel,0kids
\item \verb{[28]} 3adlts,0mar,ntrel,0kids
\item \verb{[30]} 3adlts,1mar,sp,0kids
\item \verb{[31]} 3adlts,1mar,sp,rel,0kids
\item \verb{[32]} 3adlts,1mar,rel,sp,0kids
\item \verb{[33]} 3adlts,1mar,rel,rel,0kids
\item \verb{[34]} 3adlts,1mar,sp,ntrel,0kids
\item \verb{[35]} 3adlts,1mar,ntrel,sp,0kids
\item \verb{[36]} 3adlts,1mar,rel,ntrel,0kids
\item \verb{[37]} 3adlts,1mar,ntrel,rel,0kids
\item \verb{[38]} 3adlts,1mar,ntrel,0kids
\item \verb{[40]} 3adlts,2mar,sp,0kids
\item \verb{[41]} 3adlts,2mar,sp,rel,0kids
\item \verb{[42]} 3adlts,2mar,rel,sp,0kids
\item \verb{[43]} 3adlts,2mar,rel,0kids
\item \verb{[44]} 3adlts,2mar,sp,ntrel,0kids
\item \verb{[45]} 3adlts,2mar,rel,ntrel,0kids
\item \verb{[46]} 3adlts,2mar,rel,ntrel,0kids
\item \verb{[47]} 3adlts,2mar,ntrel,rel,0kids
\item \verb{[48]} 3adlts,2mar,ntrel,0kids
\item \verb{[50]} 3adlts,3mar,sp,0kids
\item \verb{[51]} 3adlts,3mar,sp,rel,0kids
\item \verb{[52]} 3adlts,3mar,sp,rel,0kids
\item \verb{[53]} 3adlts,3mar,rel,0kids
\item \verb{[54]} 3adlts,3mar,sp,ntrel,0kids
\item \verb{[55]} 3adlts,3mar,ntrel,sp,0kids
\item \verb{[56]} 3adlts,3mar,rel,ntrel,0kids
\item \verb{[57]} 3adlts,3mar,ntrel,rel,0kids
\item \verb{[58]} 3adlts,3mar,ntrel,0kids
\item \verb{[100]} 4+adlts,0mar,0kids
\item \verb{[101]} 4+adlts,1mar,0kids
\item \verb{[102]} 4+adlts,2mar,0kids
\item \verb{[103]} 4+adlts,3mar,0kids
\item \verb{[104]} 4+adlts,4+mar,0kids
\item \verb{[120]} 3adlt,0mar,sp,1+kids
\item \verb{[121]} 3adlts,0mar,sp,rel,1+kids
\item \verb{[122]} 3adlts,0mar,rel,sp,1+kids
\item \verb{[123]} 3adlts,0mar,rel,1+kids
\item \verb{[124]} 3adlts,0mar,sp,ntrel,1+kids
\item \verb{[125]} 3adlts,0mar,ntrel,sp,1+kids
\item \verb{[126]} 3adlts,0mar,rel,1+kids
\item \verb{[127]} 3adlts,0mar,ntrel,rel,1+kids
\item \verb{[128]} 3adlts,0mar,ntrel,1+kids
\item \verb{[130]} 3adlts,1mar,sp,1+kids
\item \verb{[131]} 3adlts,1mar,sp,rel,1+kids
\item \verb{[132]} 3adlts,1mar,rel,sp,1+kids
\item \verb{[133]} 3adlts,1mar,rel,rel,1+kids
\item \verb{[134]} 3adlts,1mar,sp,ntrel,1+kids
\item \verb{[135]} 3adlts,1mar,ntrel,sp,1+kids
\item \verb{[136]} 3adlts,1mar,rel,ntrel,1+kids
\item \verb{[137]} 3adlts,1mar,ntrel,rel,1+kids
\item \verb{[138]} 3adlts,1mar,ntrel,1+kids
\item \verb{[140]} 3adlts,2mar,sp,1+kids
\item \verb{[141]} 3adlts,2mar,sp,rel,1+kids
\item \verb{[142]} 3adlts,2mar,rel,sp,1+kids
\item \verb{[143]} 3adlts,2mar,rel,1+kids
\item \verb{[144]} 3adlts,2mar,sp,ntrel,1+kids
\item \verb{[145]} 3adlts,2mar,rel,ntrel,1+kids
\item \verb{[146]} 3adlts,2mar,rel,ntrel,1+kids
\item \verb{[147]} 3adlts,2mar,ntrel,rel,1+kids
\item \verb{[148]} 3adlts,2mar,ntrel,1+kids
\item \verb{[150]} 3adlts,3mar,sp,1+kids
\item \verb{[151]} 3adlts,3mar,sp,rel,1+kids
\item \verb{[152]} 3adlts,3mar,sp,rel,1+kids
\item \verb{[153]} 3adlts,3mar,rel,1+kids
\item \verb{[154]} 3adlts,3mar,sp,ntrel,1+kids
\item \verb{[155]} 3adlts,3mar,ntrel,sp,1+kids
\item \verb{[156]} 3adlts,3mar,rel,ntrel,1+kids
\item \verb{[157]} 3adlts,3mar,ntrel,rel,1+kids
\item \verb{[158]} 3adlts,3mar,ntrel,1+kids
\item \verb{[200]} 4+adlts,0mar,1+kids
\item \verb{[201]} 4+adlts,1mar,1+kids
\item \verb{[202]} 4+adlts,2mar,1+kids
\item \verb{[203]} 4+adlts,3mar,1+kids
\item \verb{[204]} 4+adlts,4+mar,1+kids
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{famgen}},
\code{\link{hhtype1}},
\code{\link{relhh1}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}}
}
\concept{Household Members and Structure}
\keyword{variable}
