% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{hhtype1}
\alias{hhtype1}
\title{Household type (condensed)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype1
}
\details{
Question Household type (condensed)
}
\section{Values}{

\itemize{
\item \code{1} married couple, no children
\item \code{2} single parent
\item \code{3} other family, no children
\item \code{4} single adult
\item \code{5} cohabitating couple, no children
\item \code{6} non-family, no children
\item \code{8} unsure, no children
\item \code{11} married couple with children
\item \code{13} other family with children
\item \code{15} cohabitating couple with children
\item \code{16} non-family with children
\item \code{18} unsure with children
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5434/vshow}{GSS Data Explorer page for \code{hhtype1}}.
}

\seealso{
Other Household Members and Structure: 
\code{\link{famgen}},
\code{\link{hhtype}},
\code{\link{relhh1}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}}

Other Household: 
\code{\link{famgen}},
\code{\link{hhtype}},
\code{\link{relhh1}},
\code{\link{relhh2}},
\code{\link{relhh3}},
\code{\link{relhh4}},
\code{\link{rplace}},
\code{\link{rvisitor}},
\code{\link{visitors}}
}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
