% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{hhtype1_exp}
\alias{hhtype1_exp}
\title{Household type (condensed) (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype1_exp
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} married couple, no children
\item \code{2} single parent
\item \code{3} other family, no children
\item \code{4} single adult
\item \code{5} cohabitating couple, no children
\item \code{6} non-family, no children
\item \code{8} unsure, no children
\item \code{11} married couple with children
\item \code{13} other family with children
\item \code{15} cohabitating couple with children
\item \code{16} non-family with children
\item \code{18} unsure with children
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
