% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{hhtype1_exp}
\alias{hhtype1_exp}
\title{Household type (condensed) (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype1_exp
}
\details{
Question Recoded from experimental Household Composition variables ADULTS_EXP to RNRLTU18
}
\section{Values}{

\itemize{
\item \code{1} married couple, no children
\item \code{2} single parent
\item \code{3} other family, no children
\item \code{4} single adult
\item \code{5} cohabitating couple, no children
\item \code{6} non-family, no children
\item \code{8} unsure, no children
\item \code{11} married couple with children
\item \code{13} other family with children
\item \code{15} cohabitating couple with children
\item \code{16} non-family with children
\item \code{18} unsure with children
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8036/vshow}{GSS Data Explorer page for \code{hhtype1_exp}}.
}

\keyword{variable}
