% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{hhwkdis}
\alias{hhwkdis}
\title{How often r and spouse disagree sharing of hh wrk}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhwkdis
}
\details{
Question How often do you and your spouse/partner disagree about the sharing of household work?
}
\section{Values}{

\itemize{
\item \verb{[1]} several times a week
\item \verb{[2]} several times a month
\item \verb{[3]} several times a year
\item \verb{[4]} less often or rarely
\item \verb{[5]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4438/vshow}{GSS Data Explorer page for \code{hhwkdis}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab several times a week \tab several times a month \tab several times a year \tab less often or rarely \tab never \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 2058 \tab 47 \tab 75 \tab 61 \tab 268 \tab 211 \tab 22 \tab 23 \tab 2765 \cr
   Total \tab 2058 \tab 47 \tab 75 \tab 61 \tab 268 \tab 211 \tab 22 \tab 23 \tab 2765 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{deckids}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{spborn}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
