% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{hhwkfair}
\alias{hhwkfair}
\title{Sharing of hh work between r and spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhwkfair
}
\details{
Question Which of the following best applies to the sharing of household work between you and your spouse/partner?
}
\section{Values}{

\itemize{
\item \code{1} i do much more than my fair share of hh work
\item \code{2} i do a bit more than my fair share of hh work
\item \code{3} i do roughly my fair share of hh work
\item \code{4} i do a bit less than my fair share of hh work
\item \code{5} i do much less than my fair share of hh work
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4437/vshow}{GSS Data Explorer page for \code{hhwkfair}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{deckids}},
\code{\link{hhwkdis}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{spborn}},
\code{\link{sphhwork}}

Other Household: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{hhwkdis}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{sphhwork}}
}
\concept{Household}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
