% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{hinumoky}
\alias{hinumoky}
\title{Number hispanics in military-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
hinumoky
}
\details{
Question A. At the present time, about 4 percent of the army are Hispanics. All things considered, do you think there are too many Hispanics in the army, about the right number, or should there be more Hispanics in the army?
RECORD UNDER "A" BELOW.
}
\section{Values}{

\itemize{
\item \code{1} too many
\item \code{2} about right number
\item \code{3} should be more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/786/vshow}{GSS Data Explorer page for \code{hinumoky}}.
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
