% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{hispanic}
\alias{hispanic}
\title{Hispanic specified}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispanic
}
\details{
Question IF R IS FEMALE, READ LATINA; IF MALE, READ LATINO.
Are you Spanish, Hispanic, or Latino/Latina?
Yes . . . . . (ASK A)
No   . . . . . . (GO TO NEXT QUESTION)
IF YES:                  Which group are you from?
Mexican, Mexican American, Chicano/Chicana
Puerto Rican
Cuban
Other
}
\section{Values}{

\itemize{
\item \verb{[1]} not hispanic
\item \verb{[2]} mexican, mexican american, chicano/a
\item \verb{[3]} puerto rican
\item \verb{[4]} cuban
\item \verb{[5]} another hispanic, latino, or spanish origin
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5276/vshow}{GSS Data Explorer page for \code{hispanic}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab not hispanic \tab mexican, mexican american, chicano/a \tab puerto rican \tab cuban \tab another hispanic, latino, or spanish origin \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   2000 \tab - \tab 2604 \tab 110 \tab 44 \tab 6 \tab 52 \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2561 \tab 105 \tab 39 \tab 7 \tab 51 \tab 2 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2563 \tab 156 \tab 33 \tab 7 \tab 47 \tab 3 \tab 3 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 3922 \tab 370 \tab 69 \tab 23 \tab 124 \tab 2 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1782 \tab 148 \tab 24 \tab 15 \tab 50 \tab 1 \tab 3 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1803 \tab 147 \tab 27 \tab 10 \tab 51 \tab 3 \tab 3 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1701 \tab 168 \tab 22 \tab 9 \tab 70 \tab 3 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2139 \tab 241 \tab 42 \tab 20 \tab 90 \tab 3 \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2496 \tab 216 \tab 46 \tab 18 \tab 89 \tab - \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 1984 \tab 224 \tab 38 \tab 15 \tab 72 \tab 8 \tab 7 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 3544 \tab 245 \tab 51 \tab 21 \tab 137 \tab 1 \tab 2 \tab 21 \tab 10 \tab 4032 \cr
   2022 \tab - \tab 2930 \tab 311 \tab 68 \tab 33 \tab 165 \tab 17 \tab 5 \tab - \tab 15 \tab 3544 \cr
   Total \tab 0 \tab 30029 \tab 2441 \tab 503 \tab 184 \tab 998 \tab 44 \tab 29 \tab 21 \tab 25 \tab 34274 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}

Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}

Other Race: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\concept{Race}
\keyword{variable}
