% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{hispanic}
\alias{hispanic}
\title{Hispanic specified}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispanic
}
\details{
Question IF R IS FEMALE, READ LATINA; IF MALE, READ LATINO.
Are you Spanish, Hispanic, or Latino/Latina?
Yes . . . . . (ASK A)
No   . . . . . . (GO TO NEXT QUESTION)
IF YES:                  Which group are you from?
Mexican, Mexican American, Chicano/Chicana
Puerto Rican
Cuban
Other
}
\section{Values}{

\itemize{
\item \code{1} not hispanic
\item \code{2} mexican, mexican american, chicano/a
\item \code{3} puerto rican
\item \code{4} cuban
\item \code{5} another hispanic, latino, or spanish origin
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5276/vshow}{GSS Data Explorer page for \code{hispanic}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab not hispanic \tab mexican, mexican american, chicano/a \tab puerto rican \tab cuban \tab salvadorian \tab guatemalan \tab panamanian \tab central american \tab dominican \tab peruvian \tab ecuadorian \tab colombian \tab venezuelan \tab spanish \tab basque \tab filipino/a \tab latin american \tab south american \tab latin \tab latino/a \tab other, not specified \tab don't know \tab nicaraguan \tab costa rican \tab honduran \tab west indian \tab hispanic \tab argentinian \tab no answer \tab chilean \tab american \tab skipped on web \tab Total \cr
   2000 \tab - \tab 2604 \tab 110 \tab 44 \tab 6 \tab 6 \tab 1 \tab 2 \tab 1 \tab 6 \tab 4 \tab 2 \tab 4 \tab 1 \tab 9 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2561 \tab 105 \tab 39 \tab 7 \tab 5 \tab 2 \tab 2 \tab 3 \tab 5 \tab 1 \tab 2 \tab 3 \tab 3 \tab 10 \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2563 \tab 156 \tab 33 \tab 7 \tab 4 \tab 5 \tab - \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 15 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 3922 \tab 370 \tab 69 \tab 23 \tab 21 \tab 6 \tab 2 \tab 5 \tab 13 \tab 6 \tab 7 \tab 9 \tab 3 \tab 19 \tab - \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab - \tab 2 \tab 6 \tab 2 \tab 10 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1782 \tab 148 \tab 24 \tab 15 \tab 13 \tab 8 \tab 1 \tab 3 \tab 2 \tab 1 \tab 3 \tab 3 \tab - \tab 9 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1803 \tab 147 \tab 27 \tab 10 \tab 11 \tab 3 \tab 2 \tab 1 \tab 5 \tab 2 \tab 7 \tab 3 \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1701 \tab 168 \tab 22 \tab 9 \tab 7 \tab 7 \tab 3 \tab 1 \tab 10 \tab 2 \tab 3 \tab 6 \tab - \tab 15 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab - \tab 6 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2139 \tab 241 \tab 42 \tab 20 \tab 11 \tab 11 \tab - \tab 3 \tab 12 \tab 7 \tab 2 \tab 4 \tab 2 \tab 20 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab 6 \tab - \tab 3 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2496 \tab 216 \tab 46 \tab 18 \tab 13 \tab 9 \tab 2 \tab 5 \tab 9 \tab 1 \tab 4 \tab 3 \tab 1 \tab 18 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 7 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 1984 \tab 224 \tab 38 \tab 15 \tab 7 \tab 14 \tab 1 \tab 2 \tab 10 \tab - \tab 6 \tab 4 \tab 1 \tab 15 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 8 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab 7 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 21 \tab 3544 \tab 245 \tab 51 \tab 21 \tab 9 \tab 2 \tab 4 \tab 5 \tab 17 \tab 4 \tab 8 \tab 10 \tab 5 \tab 46 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab - \tab 10 \tab 4032 \cr
   2022 \tab - \tab 2930 \tab 311 \tab 68 \tab 33 \tab 22 \tab 16 \tab 6 \tab 1 \tab 13 \tab 4 \tab 7 \tab 12 \tab 8 \tab 47 \tab - \tab 2 \tab - \tab 7 \tab - \tab - \tab 2 \tab 17 \tab 4 \tab 2 \tab 6 \tab - \tab 3 \tab 3 \tab 5 \tab - \tab - \tab 15 \tab 3544 \cr
   Total \tab 21 \tab 30029 \tab 2441 \tab 503 \tab 184 \tab 129 \tab 84 \tab 25 \tab 33 \tab 103 \tab 35 \tab 53 \tab 63 \tab 27 \tab 229 \tab 2 \tab 10 \tab 5 \tab 34 \tab 3 \tab 14 \tab 15 \tab 44 \tab 25 \tab 10 \tab 50 \tab 1 \tab 26 \tab 16 \tab 29 \tab 5 \tab 1 \tab 25 \tab 34274 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\keyword{variable}
