% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{hispanic_0022}
\alias{hispanic_0022}
\title{R hispanic specific}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispanic_0022
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} not hispanic
\item \code{2} mexican, mexican american, chicano/a
\item \code{3} puerto rican
\item \code{4} cuban
\item \code{5} salvadorian
\item \code{6} guatemalan
\item \code{7} panamanian
\item \code{8} nicaraguan
\item \code{9} costa rican
\item \code{10} central american
\item \code{11} honduran
\item \code{15} dominican
\item \code{16} west indian
\item \code{20} peruvian
\item \code{21} ecuadorian
\item \code{22} colombian
\item \code{23} venezuelan
\item \code{24} argentinian
\item \code{25} chilean
\item \code{30} spanish
\item \code{31} basque
\item \code{35} filipino/a
\item \code{40} latin american
\item \code{41} south american
\item \code{42} american
\item \code{45} latin
\item \code{46} latino/a
\item \code{47} hispanic
\item \code{50} other, not specified
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
