% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{hispanic_0022}
\alias{hispanic_0022}
\title{R hispanic specific}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispanic_0022
}
\details{
Question IF R IS FEMALE, READ LATINA; IF MALE, READ LATINO.
Are you Spanish, Hispanic, or Latino/Latina?
Yes . . . . . (ASK A)
No   . . . . . . (GO TO NEXT QUESTION)
IF YES:                  Which group are you from?
Mexican, Mexican American, Chicano/Chicana
Puerto Rican
Cuban
Other
As of 2024, HISPANIC has been recoded to remove certain sensitive categories. This has been retroactively applied to all previous years of HISPANIC. The original HISPANIC containing data from 2000 to 2022 has been renamed HISPANIC_0022.
}
\section{Values}{

\itemize{
\item \code{1} not hispanic
\item \code{2} mexican, mexican american, chicano/a
\item \code{3} puerto rican
\item \code{4} cuban
\item \code{5} salvadorian
\item \code{6} guatemalan
\item \code{7} panamanian
\item \code{8} nicaraguan
\item \code{9} costa rican
\item \code{10} central american
\item \code{11} honduran
\item \code{15} dominican
\item \code{16} west indian
\item \code{20} peruvian
\item \code{21} ecuadorian
\item \code{22} colombian
\item \code{23} venezuelan
\item \code{24} argentinian
\item \code{25} chilean
\item \code{30} spanish
\item \code{31} basque
\item \code{35} filipino/a
\item \code{40} latin american
\item \code{41} south american
\item \code{42} american
\item \code{45} latin
\item \code{46} latino/a
\item \code{47} hispanic
\item \code{50} other, not specified
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8112/vshow}{GSS Data Explorer page for \code{hispanic_0022}}.
}

\keyword{variable}
