% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_259.R
\name{hispschl}
\alias{hispschl}
\title{Does r know hispanics from school}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispschl
}
\details{
Question Now, I'm going to ask some questions about these people you personally know.
A. Do you know any of these Hispanics from when you went to school or college?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2350/vshow}{GSS Data Explorer page for \code{hispschl}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1878 \tab 369 \tab 501 \tab 3 \tab 66 \tab 2817 \cr
   Total \tab 1878 \tab 369 \tab 501 \tab 3 \tab 66 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkwrk}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
