% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_257.R
\name{hispwork}
\alias{hispwork}
\title{How many people hispanic at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
hispwork
}
\details{
Question How many of the people who work where you work are Hispanic?  Are all Hispanic, almost all Hispanic, most Hispanic, about half Hispanic, most not Hispanic, almost none Hispanic, or are none Hispanic?
}
\section{Values}{

\itemize{
\item \code{1} all hispanic
\item \code{2} almost all hispanic
\item \code{3} most hispanic
\item \code{4} about half hispanic
\item \code{5} most not hispanic
\item \code{6} almost non hispanic
\item \code{7} none hispanic
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2329/vshow}{GSS Data Explorer page for \code{hispwork}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{ethchng}},
\code{\link{knwwht}}

Other Hispanics: 
\code{\link{comhisp}}
}
\concept{Hispanics}
\concept{Multi-Ethnic United States}
\keyword{variable}
