% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_376.R
\name{histsci}
\alias{histsci}
\title{How scientific: history}
\source{
General Social Survey https://gss.norc.org
}
\description{
histsci
}
\details{
Question Please look at Card B20. How scientific are each of the following fields? If you have not heard of a particular field, just say you haven't heard of it
C. History. Is history very scientific, pretty scientific, not too scientific, or not scientific at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very scientific
\item \verb{[2]} pretty scientific
\item \verb{[3]} not too scientific
\item \verb{[4]} not scientific at all
\item \verb{[5]} haven't heard of it (volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3408/vshow}{GSS Data Explorer page for \code{histsci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab haven't heard of it (volunteered) \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 195 \tab 389 \tab 676 \tab 539 \tab 6 \tab 55 \tab 4 \tab 4510 \cr
   2012 \tab 1477 \tab 62 \tab 109 \tab 203 \tab 102 \tab 1 \tab 15 \tab 5 \tab 1974 \cr
   Total \tab 4123 \tab 257 \tab 498 \tab 879 \tab 641 \tab 7 \tab 70 \tab 9 \tab 6484 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{physcsci}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{socsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
