% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{hit}
\alias{hit}
\title{Ever punched or beaten by another person}
\source{
General Social Survey https://gss.norc.org
}
\description{
hit
}
\details{
Question Have you ever been punched or beaten by another person?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1980, 1983, 1984, 1986, 1987 \cr
   B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/658/vshow}{GSS Data Explorer page for \code{hit}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 413 \tab 1090 \tab 1 \tab - \tab 1504 \cr
   1975 \tab - \tab 472 \tab 1017 \tab - \tab 1 \tab 1490 \cr
   1976 \tab - \tab 424 \tab 1075 \tab - \tab - \tab 1499 \cr
   1978 \tab - \tab 537 \tab 993 \tab 2 \tab - \tab 1532 \cr
   1980 \tab - \tab 490 \tab 977 \tab - \tab 1 \tab 1468 \cr
   1983 \tab - \tab 729 \tab 868 \tab 1 \tab 1 \tab 1599 \cr
   1984 \tab - \tab 585 \tab 885 \tab 1 \tab 2 \tab 1473 \cr
   1986 \tab - \tab 535 \tab 935 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 653 \tab 1161 \tab - \tab 5 \tab 1819 \cr
   1988 \tab 484 \tab 352 \tab 643 \tab 2 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 378 \tab 657 \tab - \tab - \tab 1537 \cr
   1990 \tab 473 \tab 329 \tab 568 \tab 1 \tab 1 \tab 1372 \cr
   1991 \tab 500 \tab 347 \tab 665 \tab - \tab 5 \tab 1517 \cr
   1993 \tab 549 \tab 379 \tab 677 \tab - \tab 1 \tab 1606 \cr
   1994 \tab 2481 \tab 195 \tab 315 \tab - \tab 1 \tab 2992 \cr
   Total \tab 4989 \tab 6818 \tab 12526 \tab 8 \tab 18 \tab 24359 \cr
}
}

\seealso{
Other Core: 
\code{\link{antiwar}},
\code{\link{civright}},
\code{\link{gun}},
\code{\link{gunage}},
\code{\link{hitage}},
\code{\link{hitnum}},
\code{\link{prowar}},
\code{\link{school}},
\code{\link{strike}}
}
\concept{Core}
\keyword{variable}
