% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{hivafraid}
\alias{hivafraid}
\title{Afraid to be around a person with hiv}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivafraid
}
\details{
Question Please indicate how much you agree or disagree with each of the following comments. I would be afraid to be around a person with HIV because I would be worried I could get infected.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly disagree
\item \verb{[2]} somewhat disagree
\item \verb{[3]} somewhat agree
\item \verb{[4]} strongly agree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7556/vshow}{GSS Data Explorer page for \code{hivafraid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly disagree \tab somewhat disagree \tab somewhat agree \tab strongly agree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1713 \tab 1249 \tab 508 \tab 270 \tab 147 \tab 126 \tab 4 \tab 15 \tab 4032 \cr
   2022 \tab 1900 \tab 792 \tab 328 \tab 273 \tab 129 \tab 80 \tab 8 \tab 34 \tab 3544 \cr
   Total \tab 3613 \tab 2041 \tab 836 \tab 543 \tab 276 \tab 206 \tab 12 \tab 49 \tab 7576 \cr
}
}

\keyword{variable}
