% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{hivimmrl}
\alias{hivimmrl}
\title{People with hiv have participated in immoral activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivimmrl
}
\details{
Question (Please indicate how much you agree or disagree with each of the following comments.) People who have HIV have participated in immoral activities.
}
\section{Values}{

\itemize{
\item \code{1} strongly disagree
\item \code{2} somewhat disagree
\item \code{3} somewhat agree
\item \code{4} strongly agree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7557/vshow}{GSS Data Explorer page for \code{hivimmrl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly disagree \tab somewhat disagree \tab somewhat agree \tab strongly agree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1713 \tab 1163 \tab 500 \tab 280 \tab 89 \tab 254 \tab 6 \tab 27 \tab 4032 \cr
   2022 \tab 1900 \tab 764 \tab 337 \tab 230 \tab 82 \tab 178 \tab 8 \tab 45 \tab 3544 \cr
   Total \tab 3613 \tab 1927 \tab 837 \tab 510 \tab 171 \tab 432 \tab 14 \tab 72 \tab 7576 \cr
}
}

\keyword{variable}
