% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{hivkiss}
\alias{hivkiss}
\title{Kiss can spread hiv}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivkiss
}
\details{
Question D. People are likely to get HIV by deep kissing, putting their tongue in their partner's mouth, if their partner has HIV.
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5080/vshow}{GSS Data Explorer page for \code{hivkiss}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2014 \tab 707 \tab 481 \tab 1210 \tab 102 \tab 38 \tab 2538 \cr
   Total \tab 707 \tab 481 \tab 1210 \tab 102 \tab 38 \tab 2538 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidscndm}},
\code{\link{hivvac}},
\code{\link{sexornt}}
}
\concept{High Risk Behaviors}
\keyword{variable}
