% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{hivtest2}
\alias{hivtest2}
\title{Where did you have your last hiv test}
\source{
General Social Survey https://gss.norc.org
}
\description{
hivtest2
}
\details{
Question Have you ever been tested for HIV? Do not count tests you may have had as part of a blood donation. Include oral test (where they take a swab from your mouth). B. Where did you have your last HIV test – at a private doctor or HMO office, at a counselling and testing site, at a hospital, at a clinic, in a jail or prison, at home, or somewhere else?
As of 2024, HIVTEST2 has been recoded to only account for seven categories by combining jail, home, and drug treatment facility into "Somewhere else." This has been retroactively applied to all previous years of HIVTEST2. The original HIVTEST2 containing data from 2006 to 2022 has been renamed HIVTEST2_0622.
}
\section{Values}{

\itemize{
\item \code{1} private doctor or hmo office
\item \code{2} counseling and testing site
\item \code{3} hospital
\item \code{4} clinic
\item \code{5} jail or prison
\item \code{6} home
\item \code{7} somewhere else
\item \code{8} emergency room
\item \code{9} drug treatment facility
\item \code{10} hospital inpatient
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5076/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab clinic \tab counseling and testing site \tab don't know \tab home \tab hospital \tab jail or prison \tab no answer \tab private doctor or hmo office \tab somewhere else \tab drug treatment facility \tab emergency room \tab hospital inpatient \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3556 \tab 185 \tab 48 \tab 5 \tab 28 \tab 145 \tab 19 \tab 18 \tab 458 \tab 48 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1295 \tab 162 \tab 41 \tab 3 \tab 14 \tab 103 \tab 13 \tab 17 \tab 325 \tab 50 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1300 \tab 149 \tab 46 \tab 2 \tab 17 \tab 108 \tab 18 \tab 24 \tab 342 \tab 38 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1249 \tab 159 \tab 25 \tab 1 \tab 9 \tab 106 \tab 11 \tab 32 \tab 343 \tab 39 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1510 \tab 227 \tab 35 \tab 3 \tab 17 \tab 187 \tab 19 \tab 38 \tab 456 \tab 46 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2135 \tab 174 \tab 39 \tab 1 \tab 8 \tab 126 \tab 11 \tab 26 \tab 312 \tab 35 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1768 \tab 119 \tab 16 \tab - \tab 6 \tab 94 \tab 10 \tab 24 \tab 290 \tab 21 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3294 \tab 123 \tab 43 \tab - \tab 9 \tab - \tab 8 \tab 21 \tab 418 \tab 55 \tab 9 \tab 17 \tab 35 \tab - \tab - \tab 4032 \cr
   2022 \tab 2971 \tab 31 \tab 31 \tab 23 \tab 13 \tab 19 \tab 113 \tab 11 \tab 283 \tab 6 \tab 38 \tab 4 \tab - \tab 1 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 65588 \tab 1329 \tab 324 \tab 38 \tab 121 \tab 888 \tab 222 \tab 211 \tab 3227 \tab 338 \tab 47 \tab 21 \tab 35 \tab 1 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{idu30}},
\code{\link{relatsex}}

Other Health: 
\code{\link{aidslook}},
\code{\link{hivtest}},
\code{\link{hivtest1}}
}
\concept{Health}
\concept{High Risk Behaviors}
\keyword{variable}
