% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{hlphitec}
\alias{hlphitec}
\title{Govt support for new products and tech}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlphitec
}
\details{
Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it.
F. Support for industry to develop new products and technology.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly in favor of
\item \verb{[2]} in favor
\item \verb{[3]} neither in favor nor against
\item \verb{[4]} against
\item \verb{[5]} strongly against
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4054/vshow}{GSS Data Explorer page for \code{hlphitec}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly in favor of \tab in favor \tab neither in favor nor against \tab against \tab strongly against \tab no answer \tab don't know \tab Total \cr
   1985 \tab 857 \tab 151 \tab 328 \tab 123 \tab 54 \tab 10 \tab 11 \tab - \tab 1534 \cr
   1990 \tab 155 \tab 306 \tab 562 \tab 241 \tab 59 \tab 18 \tab 30 \tab 1 \tab 1372 \cr
   1996 \tab 1572 \tab 324 \tab 676 \tab 203 \tab 67 \tab 18 \tab 41 \tab 3 \tab 2904 \cr
   2006 \tab 2992 \tab 550 \tab 746 \tab 129 \tab 60 \tab 19 \tab - \tab 14 \tab 4510 \cr
   2016 \tab 1477 \tab 325 \tab 764 \tab 221 \tab 43 \tab 15 \tab 9 \tab 13 \tab 2867 \cr
   Total \tab 7053 \tab 1656 \tab 3076 \tab 917 \tab 283 \tab 80 \tab 91 \tab 31 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{poleff10}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{spenviro}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
