% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_85.R
\name{hlpminr}
\alias{hlpminr}
\title{Help minorities--r's position}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlpminr
}
\details{
Question Some people feel that the government in Washington should make every possible effort to improve the social and economic position of blacks and other minority groups, even if it means giving them preferential treatment.
(Suppose these people are at one end of the scale at point number 1.)  Others feel that the government should not make any special effort to help minorities because they should help themselves. (Suppose these people are at the other end, at point 7. And of course, some other people have opinions somewhere in between at points (2, 3, 4, 5, or 6.)  ENTER CODE IN BOXES.
A. Where would you place yourself on this scale, or haven't you thought much about this?
}
\section{Values}{

\itemize{
\item \code{0} havent thought much
\item \code{1} govt help
\item \code{7} help selves
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/761/vshow}{GSS Data Explorer page for \code{hlpminr}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab havent thought much \tab govt help \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab help selves \tab don't know \tab no answer \tab Total \cr
   1980 \tab - \tab 80 \tab 68 \tab 70 \tab 139 \tab 339 \tab 242 \tab 207 \tab 296 \tab 23 \tab 4 \tab 1468 \cr
   Total \tab 0 \tab 80 \tab 68 \tab 70 \tab 139 \tab 339 \tab 242 \tab 207 \tab 296 \tab 23 \tab 4 \tab 1468 \cr
}
}

\seealso{
Other Core: 
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{hlpminfg}},
\code{\link{impchurh}},
\code{\link{impfam}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imprelax}},
\code{\link{impwork}}
}
\concept{Core}
\keyword{variable}
