% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{hlpsoc}
\alias{hlpsoc}
\title{Importance of social usefulness in a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlpsoc
}
\details{
Question On the following list there are various aspects of jobs. Please circle one number to show how important you personally consider it is in a job:
H. A job that is useful to society?
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} neither important nor unimportant
\item \code{4} not important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4514/vshow}{GSS Data Explorer page for \code{hlpsoc}}.
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
