% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_562.R
\name{hlth13}
\alias{hlth13}
\title{A death of child}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlth13
}
\details{
Question Did any deaths occur during the past year, that is since (CURRENT MONTH) 2003? Who was that?
13. Child
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5148/vshow}{GSS Data Explorer page for \code{hlth13}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 7 \tab 1316 \tab 1 \tab 16 \tab 2812 \cr
   Total \tab 1472 \tab 7 \tab 1316 \tab 1 \tab 16 \tab 2812 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth12}},
\code{\link{hlth14}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{work1}},
\code{\link{work2}},
\code{\link{work3}}
}
\concept{Negative Life Events}
\keyword{variable}
