% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_562.R
\name{hlth14}
\alias{hlth14}
\title{A death of parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlth14
}
\details{
Question Did any deaths occur during the past year, that is since (CURRENT MONTH) 2003? Who was that?
14. Father or Mother
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5149/vshow}{GSS Data Explorer page for \code{hlth14}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 35 \tab 1288 \tab 17 \tab 2812 \cr
   Total \tab 1472 \tab 35 \tab 1288 \tab 17 \tab 2812 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth10}},
\code{\link{hlth11}},
\code{\link{hlth12}},
\code{\link{hlth13}},
\code{\link{hlth8}},
\code{\link{hlth9}},
\code{\link{work1}},
\code{\link{work2}},
\code{\link{work3}}
}
\concept{Negative Life Events}
\keyword{variable}
