% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{hlthacc1}
\alias{hlthacc1}
\title{Rich vs poor access to healthcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthacc1
}
\details{
Question In the United States, do you think it is easier or harder to get access to health care... for rich people than for poor people
}
\section{Values}{

\itemize{
\item \code{1} much easier
\item \code{2} somewhat easier
\item \code{3} about the same
\item \code{4} somewhat harder
\item \code{5} much harder
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7797/vshow}{GSS Data Explorer page for \code{hlthacc1}}.

Counts by year:\tabular{llllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab Total \cr
   2022 \tab 2430 \tab 693 \tab 211 \tab 143 \tab 41 \tab 26 \tab 3544 \cr
   Total \tab 2430 \tab 693 \tab 211 \tab 143 \tab 41 \tab 26 \tab 3544 \cr
}
}

\keyword{variable}
