% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_647.R
\name{hlthacc4}
\alias{hlthacc4}
\title{Citizen vs noncitizen access to healthcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthacc4
}
\details{
Question In the United States, do you think it is easier or harder to get access to health care... for citizens of the United States than for people who do not hold U.S.citizenship
}
\section{Values}{

\itemize{
\item \code{1} much easier
\item \code{2} somewhat easier
\item \code{3} about the same
\item \code{4} somewhat harder
\item \code{5} much harder
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7800/vshow}{GSS Data Explorer page for \code{hlthacc4}}.

Counts by year:\tabular{llllllll}{
   year \tab NA \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab Total \cr
   2022 \tab 2509 \tab 319 \tab 244 \tab 236 \tab 123 \tab 113 \tab 3544 \cr
   Total \tab 2509 \tab 319 \tab 244 \tab 236 \tab 123 \tab 113 \tab 3544 \cr
}
}

\keyword{variable}
