% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_444.R
\name{hlthcare}
\alias{hlthcare}
\title{Govts resp: provide hlth care for sick}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthcare
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to . . .
C. Provide health care for the sick.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should be
\item \verb{[2]} probably should be
\item \verb{[3]} probably should not be
\item \verb{[4]} definitely should not be
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4078/vshow}{GSS Data Explorer page for \code{hlthcare}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 233 \tab 306 \tab 83 \tab 28 \tab 13 \tab 14 \tab 1534 \cr
   1990 \tab 155 \tab 464 \tab 562 \tab 92 \tab 33 \tab 27 \tab 39 \tab 1372 \cr
   1996 \tab 1572 \tab 478 \tab 572 \tab 145 \tab 46 \tab 35 \tab 56 \tab 2904 \cr
   2006 \tab 2992 \tab 850 \tab 502 \tab 116 \tab 38 \tab 11 \tab 1 \tab 4510 \cr
   2016 \tab 1477 \tab 679 \tab 493 \tab 146 \tab 49 \tab 13 \tab 10 \tab 2867 \cr
   Total \tab 7053 \tab 2704 \tab 2435 \tab 582 \tab 194 \tab 99 \tab 120 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{equalize}},
\code{\link{protstrs}},
\code{\link{racists}},
\code{\link{revoltrs}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
