% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{hlthdays}
\alias{hlthdays}
\title{Days of activity limitation past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthdays
}
\details{
Question C. During the past 30 days, for about how many days did your poor physical or mental health keep you from doing your usual activities, such as self-care, work, or recreation?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2829/vshow}{GSS Data Explorer page for \code{hlthdays}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 25 \tab 28 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab 12 \tab 27 \tab 9 \tab 24 \tab 26 \tab 29 \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 1389 \tab 80 \tab 22 \tab 5 \tab 21 \tab 1 \tab 1 \tab 1 \tab 72 \tab 11 \tab 6 \tab 1 \tab 2 \tab 1 \tab 49 \tab 24 \tab 34 \tab 34 \tab 2 \tab 16 \tab 2 \tab 6 \tab 16 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 1392 \tab 67 \tab 21 \tab 7 \tab 14 \tab 1 \tab - \tab 1 \tab 71 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab 36 \tab 18 \tab 25 \tab 29 \tab 6 \tab 13 \tab 5 \tab 2 \tab 12 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2010 \tab 857 \tab 859 \tab 53 \tab 11 \tab 7 \tab 16 \tab - \tab - \tab - \tab 52 \tab 11 \tab 1 \tab - \tab - \tab - \tab 42 \tab 14 \tab 25 \tab 39 \tab 4 \tab 11 \tab 4 \tab 6 \tab 26 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 1026 \tab 38 \tab 12 \tab 5 \tab 7 \tab - \tab - \tab - \tab 46 \tab 6 \tab - \tab - \tab 1 \tab - \tab 25 \tab 11 \tab 12 \tab 30 \tab 2 \tab 14 \tab 3 \tab - \tab 8 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 1058 \tab 65 \tab 26 \tab 6 \tab 21 \tab - \tab - \tab - \tab 75 \tab 9 \tab - \tab 1 \tab 3 \tab - \tab 38 \tab 14 \tab 19 \tab 38 \tab 3 \tab 24 \tab 4 \tab 4 \tab 34 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 1557 \tab 1326 \tab 61 \tab 51 \tab 9 \tab 41 \tab - \tab 2 \tab 1 \tab 104 \tab 26 \tab - \tab - \tab 5 \tab 1 \tab 63 \tab 36 \tab 37 \tab 89 \tab 11 \tab 37 \tab 10 \tab 14 \tab 18 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 34 \tab 3544 \cr
   Total \tab 8347 \tab 7050 \tab 364 \tab 143 \tab 39 \tab 120 \tab 2 \tab 3 \tab 3 \tab 420 \tab 69 \tab 10 \tab 3 \tab 11 \tab 2 \tab 253 \tab 117 \tab 152 \tab 259 \tab 28 \tab 115 \tab 28 \tab 32 \tab 114 \tab 11 \tab 3 \tab 10 \tab 1 \tab 3 \tab 3 \tab 34 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{backpain}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{strredpg}},
\code{\link{usedup}}
}
\concept{Quality of Working Life}
\keyword{variable}
