% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_277.R
\name{hlthdoc}
\alias{hlthdoc}
\title{R sought health information doctor or nur'se}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthdoc
}
\details{
Question If YES, please tell me if you tried to find such health information from:
D. A doctor, nurse, or other medical professional?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2520/vshow}{GSS Data Explorer page for \code{hlthdoc}}.
}

\seealso{
Other Information Society: 
\code{\link{hlthfrel}},
\code{\link{hlthmag2}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polpapr}},
\code{\link{poltv}}

Other Health: 
\code{\link{hlthfrel}},
\code{\link{hlthmag2}},
\code{\link{hlthtv}},
\code{\link{hlthwww}}
}
\concept{Health}
\concept{Information Society}
\keyword{variable}
