% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{hlthinfo}
\alias{hlthinfo}
\title{R sought health information past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthinfo
}
\details{
Question In the past year, that is, between (INSERT CURRENT MONTH) 1999/2001/2003 and (INSERT CURRENT MONTH) 2000/02–have you looked for information about a health concern or medical problem?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2516/vshow}{GSS Data Explorer page for \code{hlthinfo}}.
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{upload}}

Other Health: 
\code{\link{hlthmag1}},
\code{\link{hlthpapr}}
}
\concept{Health}
\concept{Information Society}
\keyword{variable}
