% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_277.R
\name{hlthmag2}
\alias{hlthmag2}
\title{R sought health information health magazines}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthmag2
}
\details{
Question If YES, please tell me if you tried to find such health information from:
C. Special health or medical magazine or newsletter?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2519/vshow}{GSS Data Explorer page for \code{hlthmag2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 2412 \tab 222 \tab 102 \tab 59 \tab 22 \tab - \tab - \tab 2817 \cr
   2002 \tab 2104 \tab 363 \tab 168 \tab 71 \tab 6 \tab 53 \tab - \tab 2765 \cr
   2004 \tab 2209 \tab 322 \tab 144 \tab 66 \tab 6 \tab 64 \tab 1 \tab 2812 \cr
   Total \tab 6725 \tab 907 \tab 414 \tab 196 \tab 34 \tab 117 \tab 1 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{hlthdoc}},
\code{\link{hlthfrel}},
\code{\link{hlthtv}},
\code{\link{hlthwww}},
\code{\link{polinfo}},
\code{\link{polmag1}},
\code{\link{polmag2}},
\code{\link{polpapr}},
\code{\link{poltv}}
}
\concept{Information Society}
\keyword{variable}
