% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{hlthmntl}
\alias{hlthmntl}
\title{R's mental health, mood, and ability to think}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthmntl
}
\details{
Question In general, how would you rate your mental health, including your mood and your ability to think?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7165/vshow}{GSS Data Explorer page for \code{hlthmntl}}.
}

\keyword{variable}
