% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{hlthmntl}
\alias{hlthmntl}
\title{R's mental health, mood, and ability to think}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthmntl
}
\details{
Question In general, how would you rate your mental health, including your mood and your ability to think?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7165/vshow}{GSS Data Explorer page for \code{hlthmntl}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 478 \tab 905 \tab 671 \tab 232 \tab 43 \tab 3 \tab 16 \tab - \tab 2348 \cr
   2021 \tab 380 \tab 579 \tab 1313 \tab 1147 \tab 476 \tab 122 \tab - \tab 1 \tab 14 \tab 4032 \cr
   Total \tab 380 \tab 1057 \tab 2218 \tab 1818 \tab 708 \tab 165 \tab 3 \tab 17 \tab 14 \tab 6380 \cr
}
}

\keyword{variable}
