% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{hlthpapr}
\alias{hlthpapr}
\title{R sought health information newspaper articles}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthpapr
}
\details{
Question If YES, please tell me if you tried to find such health information from:
A. Articles in a daily newspaper?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2517/vshow}{GSS Data Explorer page for \code{hlthpapr}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab Total \cr
   2000 \tab 2412 \tab 277 \tab 69 \tab 39 \tab 20 \tab - \tab 2817 \cr
   2002 \tab 2104 \tab 480 \tab 91 \tab 48 \tab 6 \tab 36 \tab 2765 \cr
   2004 \tab 2209 \tab 418 \tab 93 \tab 30 \tab 6 \tab 56 \tab 2812 \cr
   Total \tab 6725 \tab 1175 \tab 253 \tab 117 \tab 32 \tab 92 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{srcheng1}},
\code{\link{srcheng2}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
