% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{hlthphys}
\alias{hlthphys}
\title{R's physical health}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthphys
}
\details{
Question In general, how would you rate your physical health?
}
\section{Values}{

\itemize{
\item \verb{[1]} excellent
\item \verb{[2]} very good
\item \verb{[3]} good
\item \verb{[4]} fair
\item \verb{[5]} poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7164/vshow}{GSS Data Explorer page for \code{hlthphys}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab excellent \tab very good \tab good \tab fair \tab poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 378 \tab 647 \tab 777 \tab 431 \tab 96 \tab 3 \tab 16 \tab - \tab 2348 \cr
   2021 \tab 379 \tab 526 \tab 1012 \tab 1361 \tab 582 \tab 149 \tab - \tab 1 \tab 22 \tab 4032 \cr
   Total \tab 379 \tab 904 \tab 1659 \tab 2138 \tab 1013 \tab 245 \tab 3 \tab 17 \tab 22 \tab 6380 \cr
}
}

\keyword{variable}
