% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_302.R
\name{hlthplan}
\alias{hlthplan}
\title{R had medicare or medicaid}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthplan
}
\details{
Question Do you have any health insurance, including Medicare or Medicaid?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2743/vshow}{GSS Data Explorer page for \code{hlthplan}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 2387 \tab 368 \tab 7 \tab 3 \tab 2765 \cr
   Total \tab 0 \tab 2387 \tab 368 \tab 7 \tab 3 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{belvedoc}},
\code{\link{chosedoc}},
\code{\link{diffcare}},
\code{\link{fininc}},
\code{\link{fininc1}},
\code{\link{fininc2}},
\code{\link{fininc3}},
\code{\link{fininc4}},
\code{\link{insrlmts}}
}
\concept{Doctors & Patients}
\keyword{variable}
