% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_649.R
\name{hlthprb}
\alias{hlthprb}
\title{Difficulties with work or housework due to health prob}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthprb
}
\details{
Question During the past 4 weeks how often... Have you had difficulties with work or household activities because of health problems?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} seldom
\item \code{3} sometimes
\item \code{4} often
\item \code{5} very often
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7820/vshow}{GSS Data Explorer page for \code{hlthprb}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab seldom \tab sometimes \tab often \tab very often \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2381 \tab 510 \tab 257 \tab 216 \tab 72 \tab 74 \tab 24 \tab 2 \tab 8 \tab 3544 \cr
   Total \tab 2381 \tab 510 \tab 257 \tab 216 \tab 72 \tab 74 \tab 24 \tab 2 \tab 8 \tab 3544 \cr
}
}

\keyword{variable}
