% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{hlthtax}
\alias{hlthtax}
\title{Willing to pay higher taxes to imp health care for all}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthtax
}
\details{
Question How willing would you be to pay higher taxes to improve the level of health care for all people in the United States? Would you be ...
}
\section{Values}{

\itemize{
\item \verb{[1]} very willing
\item \verb{[2]} fairly willing
\item \verb{[3]} neither willing nor unwilling
\item \verb{[4]} fairly unwilling
\item \verb{[5]} very unwilling
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7794/vshow}{GSS Data Explorer page for \code{hlthtax}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very willing \tab fairly willing \tab neither willing nor unwilling \tab fairly unwilling \tab very unwilling \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2379 \tab 154 \tab 345 \tab 230 \tab 151 \tab 229 \tab 50 \tab 2 \tab 4 \tab 3544 \cr
   Total \tab 2379 \tab 154 \tab 345 \tab 230 \tab 151 \tab 229 \tab 50 \tab 2 \tab 4 \tab 3544 \cr
}
}

\keyword{variable}
