% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_648.R
\name{hlthwblif}
\alias{hlthwblif}
\title{Looked up online - healthy lifestyle}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthwblif
}
\details{
Question During the past 12 months, how often, if at all, have you used the internet to look for information on the following topics? Information on healthy lifestyle?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} seldom
\item \verb{[3]} sometimes
\item \verb{[4]} often
\item \verb{[5]} very often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7810/vshow}{GSS Data Explorer page for \code{hlthwblif}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab seldom \tab sometimes \tab often \tab very often \tab can't choose \tab skipped on web \tab Total \cr
   2022 \tab 2406 \tab 181 \tab 182 \tab 411 \tab 216 \tab 107 \tab 38 \tab 3 \tab 3544 \cr
   Total \tab 2406 \tab 181 \tab 182 \tab 411 \tab 216 \tab 107 \tab 38 \tab 3 \tab 3544 \cr
}
}

\keyword{variable}
