% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_648.R
\name{hlthweb}
\alias{hlthweb}
\title{Look up health information online in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
hlthweb
}
\details{
Question During the past 12 months, how often, if at all, did you use the internet on any device (such as computers, tablets and smartphones) to look for health or medical information for yourself or someone else?
}
\section{Values}{

\itemize{
\item \code{1} several times a day
\item \code{2} once a day
\item \code{3} several times a week
\item \code{4} several times a month
\item \code{5} several times a year
\item \code{6} never or almost never
\item \code{8} r does not have access to the internet
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7809/vshow}{GSS Data Explorer page for \code{hlthweb}}.
}

\keyword{variable}
