% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_244.R
\name{hmo2}
\alias{hmo2}
\title{Hmo prevent doctors from prescribing tests}
\source{
General Social Survey https://gss.norc.org
}
\description{
hmo2
}
\details{
Question The health care system is changing rapidly with more and more people being treated in Health Maintenance organizations (or HMOs) and in managed care plans. How much do you agree or disagree with how the HMOs and managed care plans are working?
B. They prevent doctors from prescribing tests necessary for treatment.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly disagree
\item \verb{[2]} disagree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} agree
\item \verb{[5]} strongly agree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2194/vshow}{GSS Data Explorer page for \code{hmo2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly disagree \tab disagree \tab neither agree nor disagree \tab agree \tab strongly agree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 32 \tab 301 \tab 242 \tab 455 \tab 128 \tab 223 \tab 6 \tab 2832 \cr
   Total \tab 1445 \tab 32 \tab 301 \tab 242 \tab 455 \tab 128 \tab 223 \tab 6 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc1}},
\code{\link{hmo1}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{mntldeny}},
\code{\link{mntldif}}
}
\concept{Mental Health}
\keyword{variable}
