% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{homeband}
\alias{homeband}
\title{How home is connected to www}
\source{
General Social Survey https://gss.norc.org
}
\description{
homeband
}
\details{
Question Is your home connected to the World Wide Web by a telephone line, a DSL, cable, some other way, or not at all?
}
\section{Values}{

\itemize{
\item \code{1} telephone line
\item \code{2} dsl
\item \code{3} cable
\item \code{4} satellite
\item \code{5} wireless
\item \code{6} cellphone
\item \code{7} other
\item \code{8} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2656/vshow}{GSS Data Explorer page for \code{homeband}}.
}

\seealso{
Other Information Society: 
\code{\link{huclean}},
\code{\link{knowkid3}},
\code{\link{knowkid4}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
