% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{homeyear}
\alias{homeyear}
\title{Rmost recent home you have purchased.}
\source{
General Social Survey https://gss.norc.org
}
\description{
homeyear
}
\details{
Question Tell me about the most recent home you have purchased. When did you buy this home?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1813/vshow}{GSS Data Explorer page for \code{homeyear}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 40 \tab 41 \tab 42 \tab 45 \tab 46 \tab 48 \tab 49 \tab 50 \tab 52 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 79 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 90 \tab 91 \tab 92 \tab 93 \tab 94 \tab 95 \tab 96 \tab no answer \tab Total \cr
   1996 \tab 2451 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 4 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 5 \tab 4 \tab 6 \tab 6 \tab 8 \tab 6 \tab 4 \tab 9 \tab 6 \tab 8 \tab 7 \tab 16 \tab 11 \tab 7 \tab 6 \tab 11 \tab 16 \tab 16 \tab 15 \tab 13 \tab 23 \tab 32 \tab 28 \tab 29 \tab 24 \tab 39 \tab 39 \tab 9 \tab 2 \tab 2904 \cr
   Total \tab 2451 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 4 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 5 \tab 4 \tab 6 \tab 6 \tab 8 \tab 6 \tab 4 \tab 9 \tab 6 \tab 8 \tab 7 \tab 16 \tab 11 \tab 7 \tab 6 \tab 11 \tab 16 \tab 16 \tab 15 \tab 13 \tab 23 \tab 32 \tab 28 \tab 29 \tab 24 \tab 39 \tab 39 \tab 9 \tab 2 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
