% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{homosex}
\alias{homosex}
\title{Homosexual sex relations}
\source{
General Social Survey https://gss.norc.org
}
\description{
homosex
}
\details{
Question What about sexual relations between two adults of the same sex--do you think it is always wrong, almost always wrong, wrong only sometimes, or not wrong at all?
}
\section{Values}{

\itemize{
\item \code{1} always wrong
\item \code{2} almost always wrong
\item \code{3} wrong only sometimes
\item \code{4} not wrong at all
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
   1998 \tab A/C \tab partial \cr
   2000 \tab A/C \tab partial \cr
   2002 \tab A/C \tab partial \cr
   2004 \tab A/C \tab partial \cr
   2006 \tab A/C \tab partial \cr
   2008 \tab A/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
   2012 \tab A/C \tab partial \cr
   2014 \tab A/C \tab partial \cr
   2016 \tab A/C \tab partial \cr
   2018 \tab A/C \tab partial \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
   2024 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/634/vshow}{GSS Data Explorer page for \code{homosex}}.
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{homochng}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{spdue}},
\code{\link{sppaid}},
\code{\link{teensex}},
\code{\link{xmarsex}}

Other Gays And Lesbians: 
\code{\link{homochng}}

Other Sex: 
\code{\link{homochng}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\concept{Gays And Lesbians}
\concept{Sex}
\keyword{variable}
