% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{homosex1}
\alias{homosex1}
\title{Is homosexual sex wrong?}
\source{
General Social Survey https://gss.norc.org
}
\description{
homosex1
}
\details{
Question C. And what about sexual relations between two adults of the same sex, is it. . .
}
\section{Values}{

\itemize{
\item \code{1} always wrong
\item \code{2} almost always wrong
\item \code{3} wrong only sometimes
\item \code{4} not wrong at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4632/vshow}{GSS Data Explorer page for \code{homosex1}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{taxcheat}}
}
\concept{Gays And Lesbians}
\concept{ISSP Religion}
\concept{Sex}
\keyword{variable}
