% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_664.R
\name{hompop_exp}
\alias{hompop_exp}
\title{Total people in hh (experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
hompop_exp
}
\details{
Question Calculated variable: Total people in household according to roster
As of 2024, HOMPOP_EXP has been calculated to only account for 0 to 14+ people in the household. This has been retroactively applied to all previous years of HOMPOP_EXP. The original HOMPOP_EXP containing data from in 2022 has been renamed HOMPOP_EXP_22.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8001/vshow}{GSS Data Explorer page for \code{hompop_exp}}.
}

\concept{Core}
\concept{Household}
\keyword{variable}
