% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_230.R
\name{hope1}
\alias{hope1}
\title{I could get out of a jam}
\source{
General Social Survey https://gss.norc.org
}
\description{
hope1
}
\details{
Question (HAND CARD G2)
Using the scale on the showcard, please select the response that best describes how you think about yourself right now. Please focus on yourself and what is going on in your life at this moment. Please answer each item as I read them according to the scale on the card.
If I should find myself in a jam, I could think of many ways to get out of it.
}
\section{Values}{

\itemize{
\item \code{1} definitely false
\item \code{2} mostly false
\item \code{3} somewhat false
\item \code{4} slightly false
\item \code{5} slightly true
\item \code{6} somewhat true
\item \code{7} mostly true
\item \code{8} definitely true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6203/vshow}{GSS Data Explorer page for \code{hope1}}.
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope2}},
\code{\link{hope3}},
\code{\link{hope4}},
\code{\link{lotr1}},
\code{\link{lotr2}},
\code{\link{lotr3}},
\code{\link{lotr4}},
\code{\link{lotr5}},
\code{\link{lotr6}}
}
\concept{Hope and Optimism}
\keyword{variable}
