% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{hope5}
\alias{hope5}
\title{I can think of many ways to reach my current goals}
\source{
General Social Survey https://gss.norc.org
}
\description{
hope5
}
\details{
Question (HAND CARD G2)
(Using the scale on the showcard, please select the response that best describes how you think about yourself right now.)
I can think of many ways to reach my current goals.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely false
\item \verb{[2]} mostly false
\item \verb{[3]} somewhat false
\item \verb{[4]} slightly false
\item \verb{[5]} slightly true
\item \verb{[6]} somewhat true
\item \verb{[7]} mostly true
\item \verb{[8]} definitely true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6207/vshow}{GSS Data Explorer page for \code{hope5}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab definitely false \tab mostly false \tab somewhat false \tab slightly false \tab slightly true \tab somewhat true \tab mostly true \tab definitely true \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1409 \tab 23 \tab 25 \tab 38 \tab 54 \tab 199 \tab 374 \tab 439 \tab 286 \tab 7 \tab 13 \tab 2867 \cr
   Total \tab 1409 \tab 23 \tab 25 \tab 38 \tab 54 \tab 199 \tab 374 \tab 439 \tab 286 \tab 7 \tab 13 \tab 2867 \cr
}
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope6}}
}
\concept{Hope and Optimism}
\keyword{variable}
