% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{hosrel1}
\alias{hosrel1}
\title{Rels hospitalized in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
hosrel1
}
\details{
Question A. Number of relatives hospitalized/disabled last year.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/744/vshow}{GSS Data Explorer page for \code{hosrel1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab Total \cr
   1978 \tab - \tab 878 \tab 496 \tab 120 \tab 32 \tab 6 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 864 \tab 445 \tab 117 \tab 33 \tab 4 \tab 4 \tab 1 \tab 1468 \cr
   1983 \tab - \tab 977 \tab 473 \tab 112 \tab 26 \tab 7 \tab 4 \tab - \tab 1599 \cr
   1984 \tab - \tab 944 \tab 407 \tab 92 \tab 25 \tab 3 \tab 2 \tab - \tab 1473 \cr
   Total \tab 0 \tab 3663 \tab 1821 \tab 441 \tab 116 \tab 20 \tab 10 \tab 1 \tab 6072 \cr
}
}

\seealso{
Other Core: 
\code{\link{chlddth}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel4}},
\code{\link{madeath}},
\code{\link{padeath}},
\code{\link{unrel1}},
\code{\link{unrel4}}
}
\concept{Core}
\keyword{variable}
