% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_152.R
\name{hosthome}
\alias{hosthome}
\title{Uncomf entertaining stranger's at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
hosthome
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
I. I would feel uncomfortable entertaining people I don't know well in my home.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1430/vshow}{GSS Data Explorer page for \code{hosthome}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 116 \tab 727 \tab 619 \tab 98 \tab 32 \tab 14 \tab 1606 \cr
   Total \tab 0 \tab 116 \tab 727 \tab 619 \tab 98 \tab 32 \tab 14 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{excelart}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{impfinan}}
}
\concept{Culture}
\keyword{variable}
