% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_152.R
\name{hosthome}
\alias{hosthome}
\title{Uncomf entertaining stranger's at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
hosthome
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
I. I would feel uncomfortable entertaining people I don't know well in my home.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1430/vshow}{GSS Data Explorer page for \code{hosthome}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 116 \tab 727 \tab 619 \tab 98 \tab 32 \tab 14 \tab 1606 \cr
   Total \tab 0 \tab 116 \tab 727 \tab 619 \tab 98 \tab 32 \tab 14 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{excelart}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{impfinan}}
}
\concept{Culture}
\keyword{variable}
