% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_336.R
\name{hotargus}
\alias{hotargus}
\title{Heated arguments occur in workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
hotargus
}
\details{
Question J. Heated arguments occur in my workplace.
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3007/vshow}{GSS Data Explorer page for \code{hotargus}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 76 \tab 341 \tab 516 \tab 874 \tab 18 \tab 42 \tab 2812 \cr
   Total \tab 945 \tab 76 \tab 341 \tab 516 \tab 874 \tab 18 \tab 42 \tab 2812 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{gdjobsec}},
\code{\link{lookaway}},
\code{\link{numemps}},
\code{\link{physharm}},
\code{\link{reliedon}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{treatres}}
}
\concept{Workplace Conflict}
\keyword{variable}
