% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{hotcore}
\alias{hotcore}
\title{Sci knowledge: the center of earth is very hot}
\source{
General Social Survey https://gss.norc.org
}
\description{
hotcore
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don’t know or aren’t sure, just tell me so, and we will skip to the next question. Remember true, false, or don’t know.
A. First, the center of the Earth is very hot. Is that true or false?
}
\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3334/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab false \tab no answer \tab true \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2646 \tab 306 \tab 94 \tab 2 \tab 1462 \tab - \tab 4510 \cr
   2008 \tab 518 \tab 178 \tab 83 \tab 9 \tab 1235 \tab - \tab 2023 \cr
   2010 \tab 1103 \tab 109 \tab 56 \tab 4 \tab 772 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 134 \tab 58 \tab 5 \tab 805 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 141 \tab 69 \tab 2 \tab 1027 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 149 \tab 63 \tab 7 \tab 1171 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 117 \tab 53 \tab 3 \tab 1002 \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 59730 \tab 1134 \tab 476 \tab 32 \tab 7474 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{radioact}},
\code{\link{viruses}}

Other Knowledge: 
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{radioact}},
\code{\link{viruses}}

Other Science: 
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{lasers}},
\code{\link{odds2}},
\code{\link{radioact}},
\code{\link{viruses}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
