% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{hourly}
\alias{hourly}
\title{What is your hourly wage before tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
hourly
}
\details{
Question IF HOURLY WAGE:
A. What was your hourly wage, before any taxes or deductions?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2161/vshow}{GSS Data Explorer page for \code{hourly}}.
}

\seealso{
Other Jobs: 
\code{\link{daily}},
\code{\link{monthly}},
\code{\link{trusting}},
\code{\link{weekly}},
\code{\link{yearly}}

Other Job: 
\code{\link{daily}},
\code{\link{monthly}},
\code{\link{weekly}},
\code{\link{yearly}}

Other Occupation: 
\code{\link{daily}},
\code{\link{monthly}},
\code{\link{weekly}},
\code{\link{yearly}}
}
\concept{Job}
\concept{Jobs}
\concept{Occupation}
\keyword{variable}
