% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{howfree}
\alias{howfree}
\title{How much freedom do americans have}
\source{
General Social Survey https://gss.norc.org
}
\description{
howfree
}
\details{
Question How much freedom do you think Americans have today?  Would you say they have complete freedom, a great deal of freedom, a moderate degree of freedom, not much freedom, or no freedom at all?
}
\section{Values}{

\itemize{
\item \code{1} complete freedom
\item \code{2} a great deal of freedom
\item \code{3} a moderate degree of freedom
\item \code{4} not much freedom
\item \code{5} no freedom at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2663/vshow}{GSS Data Explorer page for \code{howfree}}.
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{freenow}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}}
}
\concept{Freedom}
\keyword{variable}
