% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{howpaid}
\alias{howpaid}
\title{How get paid last week}
\source{
General Social Survey https://gss.norc.org
}
\description{
howpaid
}
\details{
Question As of last week, did your main job pay an hourly wage, daily wage, weekly wage, monthly salary, annual salary, or other pay rate (e.g. piece rate, commission, wage + tips, and other mixed payment methods), irrespective of the time between payrolls and the use of bonuses or profit-sharing?
}
\section{Values}{

\itemize{
\item \verb{[1]} hourly wage
\item \verb{[2]} daily wage
\item \verb{[3]} weekly wage
\item \verb{[4]} monthly salary
\item \verb{[5]} annual salary
\item \verb{[6]} other pay rate
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2160/vshow}{GSS Data Explorer page for \code{howpaid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab hourly wage \tab daily wage \tab weekly wage \tab monthly salary \tab annual salary \tab other pay rate \tab no answer \tab Total \cr
   1998 \tab 1455 \tab 650 \tab 17 \tab 124 \tab 149 \tab 225 \tab 182 \tab 30 \tab 2832 \cr
   Total \tab 1455 \tab 650 \tab 17 \tab 124 \tab 149 \tab 225 \tab 182 \tab 30 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{numemp}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
