% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_564.R
\name{hrdshp4}
\alias{hrdshp4}
\title{Being unable to purchase needed food}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrdshp4
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
D. Now I am going to ask about specific hardships. Did any of the following occur to you since (CURRENT MONTH), (1990/2003)?
4. Unable to purchase needed food.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5168/vshow}{GSS Data Explorer page for \code{hrdshp4}}.
}

\seealso{
Other Negative Life Events: 
\code{\link{famper1}},
\code{\link{finan4}},
\code{\link{finan5}},
\code{\link{hrdshp1}},
\code{\link{hrdshp2}},
\code{\link{hrdshp3}},
\code{\link{hrdshp5}},
\code{\link{hrdshp6}},
\code{\link{hrdshp7}}
}
\concept{Negative Life Events}
\keyword{variable}
