% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{hrfambiz}
\alias{hrfambiz}
\title{How many hours per week?}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrfambiz
}
\details{
Question How many hours per week?
}
\section{Values}{

\itemize{
\item \verb{[1]} only especially busy seasons
\item \verb{[2]} less than 10 hours a week
\item \verb{[3]} between 10 and 30 hours a week
\item \verb{[4]} more than 30 hours a week
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1877/vshow}{GSS Data Explorer page for \code{hrfambiz}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab only especially busy seasons \tab less than 10 hours a week \tab between 10 and 30 hours a week \tab more than 30 hours a week \tab no answer \tab Total \cr
   1996 \tab 2841 \tab 8 \tab 7 \tab 12 \tab 27 \tab 9 \tab 2904 \cr
   Total \tab 2841 \tab 8 \tab 7 \tab 12 \tab 27 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
