% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{hrfambiz}
\alias{hrfambiz}
\title{How many hours per week?}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrfambiz
}
\details{
Question How many hours per week?
}
\section{Values}{

\itemize{
\item \code{1} only especially busy seasons
\item \code{2} less than 10 hours a week
\item \code{3} between 10 and 30 hours a week
\item \code{4} more than 30 hours a week
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1877/vshow}{GSS Data Explorer page for \code{hrfambiz}}.
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}

Other Labor Force Status: 
\code{\link{wkfambiz}}
}
\concept{Labor Force Status}
\concept{Markets}
\keyword{variable}
