% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_203.R
\name{hrsenvir}
\alias{hrsenvir}
\title{How many hours spent in environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrsenvir
}
\details{
Question I'd like your best estimate of the total number of hours you spent in the past month on each of the areas in which you have been a volunteer. First, how many hours did you spend in the past month working for (ASK FOR EACH ITEM R MENTIONED IN MONHLTH-MONOTH) (RECORD ABOVE UNDER MONHLTH-MONOTH TO NEAREST HALF HOUR. If UNSURE, PROBE FOR BEST ESTIMATE.)
E. Environment
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1919/vshow}{GSS Data Explorer page for \code{hrsenvir}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{hrseduc}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}}

Other Voluntary Associations: 
\code{\link{hrseduc}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\concept{Voluntary Associations}
\keyword{variable}
