% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_371.R
\name{hsbio}
\alias{hsbio}
\title{R ever took a high school biology course}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsbio
}
\details{
Question Did you take a high school biology course?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3352/vshow}{GSS Data Explorer page for \code{hsbio}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2690 \tab 1434 \tab 345 \tab 9 \tab 32 \tab 4510 \cr
   2008 \tab 597 \tab 1113 \tab 300 \tab 11 \tab 2 \tab 2023 \cr
   2010 \tab 1370 \tab 496 \tab 107 \tab 7 \tab 64 \tab 2044 \cr
   2012 \tab 1068 \tab 726 \tab 160 \tab 17 \tab 3 \tab 1974 \cr
   2014 \tab 1348 \tab 910 \tab 260 \tab 18 \tab 2 \tab 2538 \cr
   2016 \tab 1532 \tab 1050 \tab 262 \tab 15 \tab 8 \tab 2867 \cr
   2018 \tab 1222 \tab 885 \tab 222 \tab 15 \tab 4 \tab 2348 \cr
   Total \tab 9827 \tab 6614 \tab 1656 \tab 92 \tab 115 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{hschem}},
\code{\link{hsmath}},
\code{\link{hsphys}},
\code{\link{sciagrgw}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
