% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{hsclass2}
\alias{hsclass2}
\title{R 2nd favorite subject in hs}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsclass2
}
\details{
Question B. When you were in high school/In high school), what (was/is) your favorite course or subject?
}
\section{Values}{

\itemize{
\item \code{0} none
\item \code{1} english, literature
\item \code{2} foreign language
\item \code{3} fine arts
\item \code{4} other humanities
\item \code{5} economics
\item \code{6} psychology
\item \code{7} government-civics
\item \code{8} social science-other social study
\item \code{9} mathematics
\item \code{10} science
\item \code{11} other
\item \code{12} business
\item \code{13} education
\item \code{14} engineering
\item \code{15} other vocational
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1453/vshow}{GSS Data Explorer page for \code{hsclass2}}.
}

\seealso{
Other Culture: 
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{hsclass1}},
\code{\link{mostimp4}}
}
\concept{Culture}
\keyword{variable}
