% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_438.R
\name{hsjudge}
\alias{hsjudge}
\title{School subjects: good judgment}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsjudge
}
\details{
Question Here are some things that might be taught in school. How important is it that schools teach each of these to 15-year-olds?
E. Ability to make one''s own judgments.
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4019/vshow}{GSS Data Explorer page for \code{hsjudge}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidavg}},
\code{\link{aidneedy}},
\code{\link{aidsmart}},
\code{\link{colop}},
\code{\link{hscaring}},
\code{\link{hsjobtr}},
\code{\link{hslibart}},
\code{\link{hsorder}},
\code{\link{hssci}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
