% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_371.R
\name{hsmath}
\alias{hsmath}
\title{The highest level of math r completed in high school}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsmath
}
\details{
Question Now, let me ask you to think about the courses you took in high school. What was the highest level of math that you completed in high school?
}
\section{Values}{

\itemize{
\item \code{0} no math in hs; didn't go to hs
\item \code{1} general math, business, or vocational math
\item \code{2} pre-algebra
\item \code{3} one year of algebra
\item \code{4} two years of algebra
\item \code{5} geometry (plane or solid or both)
\item \code{6} trigonometry/linear programming/analysis
\item \code{7} pre-calculus
\item \code{8} calculus
\item \code{9} statistics/probability
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3351/vshow}{GSS Data Explorer page for \code{hsmath}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsphys}},
\code{\link{sciagrgw}}

Other Education: 
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsphys}}
}
\concept{Education}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
