% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_173.R
\name{hspasn10}
\alias{hspasn10}
\title{Change in ethnic status in local community}
\source{
General Social Survey https://gss.norc.org
}
\description{
hspasn10
}
\details{
Question Compared to 10 years ago do Hispanics and Asians make up a larger share of the population in your local community, the same share, or a smaller share?
}
\section{Values}{

\itemize{
\item \code{1} larger share
\item \code{2} same share
\item \code{3} smaller share
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1612/vshow}{GSS Data Explorer page for \code{hspasn10}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab larger share \tab same share \tab smaller share \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 728 \tab 408 \tab 172 \tab 157 \tab 9 \tab 2992 \cr
   Total \tab 1518 \tab 728 \tab 408 \tab 172 \tab 157 \tab 9 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}}
}
\concept{Multiculturalism}
\keyword{variable}
