% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_133.R
\name{hsphouse}
\alias{hsphouse}
\title{Hispanics discriminated in housing}
\source{
General Social Survey https://gss.norc.org
}
\description{
hsphouse
}
\details{
Question How much discrimination is there that makes it hard for Hispanic Americans to buy or rent housing wherever they want?
A. Hispanic Americans
}
\section{Values}{

\itemize{
\item \code{1} a lot
\item \code{2} some
\item \code{3} only a little
\item \code{4} none at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1246/vshow}{GSS Data Explorer page for \code{hsphouse}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influhsp}},
\code{\link{influso}},
\code{\link{racwork}}

Other Hispanics: 
\code{\link{hspjobs}},
\code{\link{influhsp}}

Other Housing: 
\code{\link{asnhouse}},
\code{\link{blkhouse}}
}
\concept{Hispanics}
\concept{Housing}
\concept{Intergroup Relations}
\keyword{variable}
