% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_262.R
\name{hswht}
\alias{hswht}
\title{High school mostly white}
\source{
General Social Survey https://gss.norc.org
}
\description{
hswht
}
\details{
Question IF ATTENDED HIGH SCHOOL (EDUC), ASK IN TERMS OF "HIGH SCHOOL."
IF LEFT SCHOOL beFORE HIGH SCHOOL, ASK IN TERMS OF "THE LAST SCHOOL YOU ATTENDED."
A. Was the high school/last school you attended all white, mostly white, about half white and half other races, mostly other races, or all other races?
}
\section{Values}{

\itemize{
\item \verb{[1]} all white
\item \verb{[2]} mostly white
\item \verb{[3]} about half white and half other races
\item \verb{[4]} mostly other races
\item \verb{[5]} all other races
\item \verb{[6]} not all white
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2376/vshow}{GSS Data Explorer page for \code{hswht}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all white \tab mostly white \tab about half white and half other races \tab mostly other races \tab all other races \tab not all white \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 286 \tab 444 \tab 159 \tab 45 \tab 55 \tab 358 \tab 51 \tab 2817 \cr
   Total \tab 1419 \tab 286 \tab 444 \tab 159 \tab 45 \tab 55 \tab 358 \tab 51 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{hseth}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
