% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{huaddwhy}
\alias{huaddwhy}
\title{Why there is no hu at this address}
\source{
General Social Survey https://gss.norc.org
}
\description{
huaddwhy
}
\details{
Question A. SELECT ONE FROM THE LIST BELOW WHICH BEST DESCRIBES WHY THERE IS NO HU AT THIS ADDRESS
}
\section{Values}{

\itemize{
\item \code{1} condemned
\item \code{2} demolished
\item \code{3} place of business without residence
\item \code{4} no such address, no such hu
\item \code{5} group quarters
\item \code{6} vacation cabin
\item \code{7} not usable as a permanent residence
\item \code{8} transient use (less than one month)
\item \code{9} not an hu for other reason
\item \code{10} under construction
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018 \cr
   B/-/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6653/vshow}{GSS Data Explorer page for \code{huaddwhy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no such address, no such hu \tab vacation cabin \tab not usable as a permanent residence \tab not an hu for other reason \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab 2840 \tab 20 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2342 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 2348 \cr
   2022 \tab 3543 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3544 \cr
   Total \tab 8725 \tab 25 \tab 1 \tab 2 \tab 5 \tab 1 \tab 0 \tab 8759 \cr
}
}

\keyword{variable}
