% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{huclean}
\alias{huclean}
\title{R hu interior cleanliness rating}
\source{
General Social Survey https://gss.norc.org
}
\description{
huclean
}
\details{
Question A. How clean was the interior of the housing unit?
}
\section{Values}{

\itemize{
\item \verb{[1]} very clean
\item \verb{[2]} clean
\item \verb{[3]} so-so
\item \verb{[4]} not very clean
\item \verb{[5]} dirty
\item \verb{[6]} interviewer never saw interior of the respondent's hu
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2655/vshow}{GSS Data Explorer page for \code{huclean}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab very clean \tab clean \tab so-so \tab not very clean \tab dirty \tab don't know \tab no answer \tab interviewer never saw interior of the respondent's hu \tab not available in this year \tab Total \cr
   2000 \tab - \tab 1169 \tab 733 \tab 330 \tab 108 \tab 70 \tab 346 \tab 61 \tab - \tab - \tab 2817 \cr
   2016 \tab - \tab 705 \tab 951 \tab 337 \tab 124 \tab 60 \tab 36 \tab 207 \tab 447 \tab - \tab 2867 \cr
   2018 \tab - \tab 566 \tab 684 \tab 263 \tab 84 \tab 53 \tab 18 \tab 295 \tab 385 \tab - \tab 2348 \cr
   2022 \tab 1864 \tab 370 \tab 452 \tab 246 \tab 82 \tab 34 \tab 13 \tab - \tab 483 \tab - \tab 3544 \cr
   Total \tab 1864 \tab 2810 \tab 2820 \tab 1176 \tab 398 \tab 217 \tab 413 \tab 563 \tab 1315 \tab 0 \tab 11576 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{homeband}},
\code{\link{knowkid3}},
\code{\link{knowkid4}}
}
\concept{Information Society}
\concept{Paradata}
\keyword{variable}
