% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_293.R
\name{huclean}
\alias{huclean}
\title{R hu interior cleanliness rating}
\source{
General Social Survey https://gss.norc.org
}
\description{
huclean
}
\details{
Question A. How clean was the interior of the housing unit?
}
\section{Values}{

\itemize{
\item \code{1} very clean
\item \code{2} clean
\item \code{3} so-so
\item \code{4} not very clean
\item \code{5} dirty
\item \code{6} interviewer never saw interior of the respondent's hu
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2655/vshow}{GSS Data Explorer page for \code{huclean}}.
}

\seealso{
Other Information Society: 
\code{\link{homeband}},
\code{\link{knowkid3}},
\code{\link{knowkid4}}
}
\concept{Dwelling (R's)}
\concept{Information Society}
\concept{Paradata}
\keyword{variable}
