% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{hunt1}
\alias{hunt1}
\title{Does r or spouse or partner hunt}
\source{
General Social Survey https://gss.norc.org
}
\description{
hunt1
}
\details{
Question Do you (or does your (husband/wife/partner)) go hunting?
}
\section{Values}{

\itemize{
\item \code{1} yes, respondent does
\item \code{2} yes, spouse or partner does
\item \code{3} yes, both do
\item \code{4} no, neither r nor spouse or partner does
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
   A/C/- \tab 2016, 2018, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6195/vshow}{GSS Data Explorer page for \code{hunt1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab yes, respondent does \tab yes, spouse or partner does \tab yes, both do \tab no, neither r nor spouse or partner does \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2016 \tab 978 \tab 184 \tab 79 \tab 57 \tab 1566 \tab 1 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 164 \tab 66 \tab 36 \tab 1307 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 325 \tab 155 \tab 93 \tab 3452 \tab - \tab - \tab 7 \tab 4032 \cr
   2022 \tab 1203 \tab 235 \tab 68 \tab 48 \tab 1985 \tab 3 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 2955 \tab 908 \tab 368 \tab 234 \tab 8310 \tab 5 \tab 3 \tab 8 \tab 12791 \cr
}
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}},
\code{\link{tvhours}}
}
\concept{Core}
\keyword{variable}
