% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{huntothr}
\alias{huntothr}
\title{Does other household member hunt}
\source{
General Social Survey https://gss.norc.org
}
\description{
huntothr
}
\details{
Question B. Does any other member of this household go hunting?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1980, 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/689/vshow}{GSS Data Explorer page for \code{huntothr}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab 738 \tab 176 \tab 497 \tab 1 \tab 118 \tab 1530 \cr
   1980 \tab 782 \tab 134 \tab 528 \tab 4 \tab 20 \tab 1468 \cr
   1982 \tab 963 \tab 143 \tab 568 \tab 4 \tab 182 \tab 1860 \cr
   1984 \tab 778 \tab 137 \tab 494 \tab 3 \tab 61 \tab 1473 \cr
   Total \tab 3261 \tab 590 \tab 2087 \tab 12 \tab 381 \tab 6331 \cr
}
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}},
\code{\link{tvhours}}
}
\concept{Core}
\keyword{variable}
