% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{hurtoth}
\alias{hurtoth}
\title{How likely x violent toward other people}
\source{
General Social Survey https://gss.norc.org
}
\description{
hurtoth
}
\details{
Question In your opinion, how likely is it NAME would do something violent toward other people? Is it:
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2009/vshow}{GSS Data Explorer page for \code{hurtoth}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 231 \tab 487 \tab 427 \tab 187 \tab 91 \tab 21 \tab 2904 \cr
   2006 \tab 3073 \tab 143 \tab 469 \tab 554 \tab 180 \tab 76 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 133 \tab 415 \tab 419 \tab 130 \tab 46 \tab 30 \tab 2348 \cr
   Total \tab 5708 \tab 507 \tab 1371 \tab 1400 \tab 497 \tab 213 \tab 66 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
