% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{hurtself}
\alias{hurtself}
\title{How likely x violent toward oneself}
\source{
General Social Survey https://gss.norc.org
}
\description{
hurtself
}
\details{
Question In your opinion, how likely is it NAME would do something violent toward him/herself? Is it:
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2010/vshow}{GSS Data Explorer page for \code{hurtself}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 408 \tab 585 \tab 256 \tab 111 \tab 67 \tab 17 \tab 2904 \cr
   2006 \tab 3073 \tab 313 \tab 610 \tab 322 \tab 105 \tab 72 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 258 \tab 539 \tab 243 \tab 59 \tab 44 \tab 30 \tab 2348 \cr
   Total \tab 5708 \tab 979 \tab 1734 \tab 821 \tab 275 \tab 183 \tab 62 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
