% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_431.R
\name{hushrel1}
\alias{hushrel1}
\title{Leader1 tried to keep r from telling other people about the relationship}
\source{
General Social Survey https://gss.norc.org
}
\description{
hushrel1
}
\details{
Question A. Did this leader say or do anything to try to keep you from telling other people about the relationship?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3957/vshow}{GSS Data Explorer page for \code{hushrel1}}.
}

\seealso{
Other Clergy/Sex: 
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}}

Other Religion: 
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}}

Other Sex: 
\code{\link{clrggen1}},
\code{\link{clrggen2}},
\code{\link{clrggen3}},
\code{\link{clrgsex1}},
\code{\link{clrgsex2}},
\code{\link{clrgsex3}},
\code{\link{ongorel1}},
\code{\link{ongorel2}},
\code{\link{ongorel3}}
}
\concept{Clergy/Sex}
\concept{Religion}
\concept{Sex}
\keyword{variable}
