% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{hvylift}
\alias{hvylift}
\title{R do repeated lifting}
\source{
General Social Survey https://gss.norc.org
}
\description{
hvylift
}
\details{
Question Does your job require you to do repeated lifting, pushing, pulling or bending?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2812/vshow}{GSS Data Explorer page for \code{hvylift}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 821 \tab 957 \tab 5 \tab 13 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 792 \tab 925 \tab 3 \tab 14 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 547 \tab 615 \tab 2 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 600 \tab 642 \tab 1 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 722 \tab 690 \tab 5 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 900 \tab 1057 \tab 5 \tab 12 \tab 15 \tab 3544 \cr
   Total \tab 8345 \tab 4382 \tab 4886 \tab 21 \tab 100 \tab 15 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
