% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{hwduties}
\alias{hwduties}
\title{Is r mainly respons.for hh chores?}
\source{
General Social Survey https://gss.norc.org
}
\description{
hwduties
}
\details{
Question Are you the person who is responsible for doing the general domestic duties - like cleaning, cooking, washing and so on - in your household?
}
\section{Values}{

\itemize{
\item \code{1} yes i responsible
\item \code{2} yes me and other
\item \code{3} no someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4495/vshow}{GSS Data Explorer page for \code{hwduties}}.
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{paysame}},
\code{\link{sexemp}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}},
\code{\link{yrsfirm}}
}
\concept{Household}
\concept{Work Orientation}
\keyword{variable}
