% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_379.R
\name{icecaps}
\alias{icecaps}
\title{Polar ice caps got larger or smaller over last 25 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
icecaps
}
\details{
Question The next few questions are about the Arctic and the Antarctic. The Arctic is the region around the North Pole; Antarctic is the region that contains the South Pole. These questions are like ones you might see on a television game show. If you don't know or aren't sure, just tell me so, and we will skip to the next question. Remember true, false, or don't know.
E. Would you say the polar ice caps have gotten larger or smaller over the last 25 years?
}
\section{Values}{

\itemize{
\item \verb{[1]} larger
\item \verb{[2]} smaller
\item \verb{[3]} stayed the same:volunteered
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3436/vshow}{GSS Data Explorer page for \code{icecaps}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab larger \tab smaller \tab stayed the same:volunteered \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 83 \tab 1408 \tab 24 \tab 348 \tab 1 \tab 4510 \cr
   2010 \tab 1338 \tab 34 \tab 563 \tab 12 \tab 86 \tab 11 \tab 2044 \cr
   Total \tab 3984 \tab 117 \tab 1971 \tab 36 \tab 434 \tab 12 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{comorsci}},
\code{\link{extinct}},
\code{\link{huntbear}},
\code{\link{icesheet}},
\code{\link{inuit}},
\code{\link{nosun}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
