% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{idecide}
\alias{idecide}
\title{My job allows me make part decisions}
\source{
General Social Survey https://gss.norc.org
}
\description{
idecide
}
\details{
Question How true would you say each of the following statements is about your job?  Would you say very true, true, somewhat true, or not true at all:
C. My job allows me to take part in making decisions that affect my work.
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} true
\item \code{3} somewhat true
\item \code{4} not true at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1310/vshow}{GSS Data Explorer page for \code{idecide}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very true \tab true \tab somewhat true \tab not true at all \tab can't choose \tab no answer \tab Total \cr
   1991 \tab 605 \tab 321 \tab 335 \tab 132 \tab 112 \tab 2 \tab 10 \tab 1517 \cr
   Total \tab 605 \tab 321 \tab 335 \tab 132 \tab 112 \tab 2 \tab 10 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{totsup}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
