% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{idu30}
\alias{idu30}
\title{R inject drugs in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
idu30
}
\details{
Question How long has it been since you last used a needle to inject drugs?
}
\section{Values}{

\itemize{
\item \code{1} within the past 30 days
\item \code{2} more than 30 days ago but within the past 12 months
\item \code{3} more than 12 months ago but within the past 3 years
\item \code{4} more than 3 years ago
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5071/vshow}{GSS Data Explorer page for \code{idu30}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab within the past 30 days \tab more than 30 days ago but within the past 12 months \tab more than 12 months ago but within the past 3 years \tab more than 3 years ago \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 2682 \tab 3 \tab 9 \tab 7 \tab 37 \tab 4 \tab 75 \tab - \tab 2817 \cr
   2002 \tab 2679 \tab 5 \tab 2 \tab 2 \tab 42 \tab 2 \tab 33 \tab - \tab 2765 \cr
   2004 \tab 2741 \tab 3 \tab 1 \tab 4 \tab 44 \tab - \tab 19 \tab - \tab 2812 \cr
   2006 \tab 4388 \tab 8 \tab 6 \tab 5 \tab 77 \tab - \tab 26 \tab - \tab 4510 \cr
   2008 \tab 1947 \tab 8 \tab 2 \tab 4 \tab 42 \tab 1 \tab 19 \tab - \tab 2023 \cr
   2010 \tab 1956 \tab 4 \tab 2 \tab 7 \tab 47 \tab 2 \tab 26 \tab - \tab 2044 \cr
   2012 \tab 1901 \tab 4 \tab 2 \tab 2 \tab 33 \tab - \tab 32 \tab - \tab 1974 \cr
   2014 \tab 2436 \tab 5 \tab 7 \tab 9 \tab 40 \tab - \tab 41 \tab - \tab 2538 \cr
   2016 \tab 2790 \tab 6 \tab 1 \tab 7 \tab 43 \tab 1 \tab 19 \tab - \tab 2867 \cr
   2018 \tab 2291 \tab 4 \tab 4 \tab 5 \tab 27 \tab - \tab 17 \tab - \tab 2348 \cr
   2021 \tab 3969 \tab 4 \tab 8 \tab 6 \tab 37 \tab 3 \tab 2 \tab 3 \tab 4032 \cr
   2022 \tab 3493 \tab 2 \tab 2 \tab 6 \tab 29 \tab 4 \tab 7 \tab 1 \tab 3544 \cr
   Total \tab 33273 \tab 56 \tab 46 \tab 64 \tab 498 \tab 17 \tab 316 \tab 4 \tab 34274 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest2}},
\code{\link{relatsex}}
}
\concept{High Risk Behaviors}
\keyword{variable}
