% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{if08who}
\alias{if08who}
\title{Who you would have voted for}
\source{
General Social Survey https://gss.norc.org
}
\description{
if08who
}
\details{
Question K. In 2008, you remember that Obama ran for President on the Democratic ticket against McCain for the Republicans. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:
Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} obama
\item \verb{[2]} mccain
\item \verb{[3]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/174/vshow}{GSS Data Explorer page for \code{if08who}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab obama \tab mccain \tab other \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1393 \tab 353 \tab 113 \tab 118 \tab 45 \tab 22 \tab 2044 \cr
   2012 \tab 1316 \tab 354 \tab 146 \tab 84 \tab 51 \tab 23 \tab 1974 \cr
   2014 \tab 1675 \tab 402 \tab 200 \tab 159 \tab 62 \tab 40 \tab 2538 \cr
   Total \tab 4384 \tab 1109 \tab 459 \tab 361 \tab 158 \tab 85 \tab 6556 \cr
}
}

\seealso{
Other Core: 
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{pres12}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
