% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{if12who}
\alias{if12who}
\title{Who would r have voted for in 2012 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if12who
}
\details{
Question L. In 2012, you remember that Obama ran for President on the Democratic ticket against Romney for the Republicans. Do you remember for sure whether or not you voted in that election?
2. Who would you have voted for, for president, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} obama
\item \verb{[2]} romney
\item \verb{[3]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/177/vshow}{GSS Data Explorer page for \code{if12who}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab obama \tab romney \tab other \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1664 \tab 372 \tab 210 \tab 202 \tab 70 \tab 20 \tab 2538 \cr
   2016 \tab 1794 \tab 532 \tab 219 \tab 154 \tab 93 \tab 75 \tab 2867 \cr
   2018 \tab 1480 \tab 500 \tab 126 \tab 110 \tab 72 \tab 60 \tab 2348 \cr
   Total \tab 4938 \tab 1404 \tab 555 \tab 466 \tab 235 \tab 155 \tab 7753 \cr
}
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{pres12}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
