% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{if16who}
\alias{if16who}
\title{Who you would have voted for}
\source{
General Social Survey https://gss.norc.org
}
\description{
if16who
}
\details{
Question Who would you have voted for, for president, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} clinton
\item \verb{[2]} trump
\item \verb{[3]} other
\item \verb{[4]} don't know/can't remember
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7091/vshow}{GSS Data Explorer page for \code{if16who}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab clinton \tab trump \tab other \tab don't know \tab no answer \tab skipped on web \tab don't know/can't remember \tab Total \cr
   2018 \tab 1503 \tab 348 \tab 163 \tab 229 \tab 54 \tab 51 \tab - \tab - \tab 2348 \cr
   2021 \tab 2842 \tab 310 \tab 248 \tab 203 \tab 83 \tab 337 \tab 9 \tab - \tab 4032 \cr
   2022 \tab 2203 \tab 365 \tab 326 \tab 170 \tab 27 \tab 305 \tab 5 \tab 143 \tab 3544 \cr
   Total \tab 6548 \tab 1023 \tab 737 \tab 602 \tab 164 \tab 693 \tab 14 \tab 143 \tab 9924 \cr
}
}

\keyword{variable}
