% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{if68who}
\alias{if68who}
\title{Who would r have voted for-1968 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if68who
}
\details{
Question A. Now in 1968, you remember that Humphrey ran for President on the Democratic ticket against Nixon for the Republicans, and Wallace as an Independent. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE: Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} humphrey
\item \verb{[2]} nixon
\item \verb{[3]} wallace
\item \verb{[4]} other
\item \verb{[5]} refused
\item \verb{[6]} wldnt vt-relig
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972, 1973 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/144/vshow}{GSS Data Explorer page for \code{if68who}}.

Counts by year:\tabular{llllllllll}{
   year \tab humphrey \tab nixon \tab wallace \tab other \tab wldnt vt-relig \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 173 \tab 175 \tab 49 \tab 23 \tab 10 \tab 88 \tab 1089 \tab 6 \tab 1613 \cr
   1973 \tab 178 \tab 196 \tab 94 \tab 14 \tab - \tab 43 \tab 961 \tab 18 \tab 1504 \cr
   Total \tab 351 \tab 371 \tab 143 \tab 37 \tab 10 \tab 131 \tab 2050 \tab 24 \tab 3117 \cr
}
}

\seealso{
Other Core: 
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
