% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{if76who}
\alias{if76who}
\title{Who would r have voted for-1976 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if76who
}
\details{
Question C. In 1976, you remember that Carter ran for President on the Democratic ticket against Ford for the Republicans. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} carter
\item \verb{[2]} ford
\item \verb{[3]} other
\item \verb{[4]} refused
\item \verb{[5]} wouldnt vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1978, 1980, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/150/vshow}{GSS Data Explorer page for \code{if76who}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab carter \tab ford \tab other \tab wouldnt vote \tab don't know \tab no answer \tab refused \tab Total \cr
   1977 \tab 988 \tab 307 \tab 151 \tab 13 \tab 16 \tab 35 \tab 20 \tab - \tab 1530 \cr
   1978 \tab 974 \tab 267 \tab 169 \tab 25 \tab 4 \tab 55 \tab 38 \tab - \tab 1532 \cr
   1980 \tab 909 \tab 243 \tab 198 \tab 28 \tab - \tab 77 \tab 13 \tab - \tab 1468 \cr
   1982 \tab 1171 \tab 341 \tab 227 \tab 33 \tab - \tab 79 \tab 8 \tab 1 \tab 1860 \cr
   Total \tab 4042 \tab 1158 \tab 745 \tab 99 \tab 20 \tab 246 \tab 79 \tab 1 \tab 6390 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
