% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{if80who}
\alias{if80who}
\title{Who would r have voted for-1980 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if80who
}
\details{
Question D. In 1980, you remember that Carter ran for President on the Democratic ticket against Reagan for the Republicans, and  anderson as an Independent.
Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} carter
\item \code{2} reagan
\item \code{3} anderson
\item \code{4} other
\item \code{5} wouldnt vote
\item \code{7} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
   1985 \tab - \tab full \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/153/vshow}{GSS Data Explorer page for \code{if80who}}.
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}

Other Political: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\concept{Political}
\keyword{variable}
