% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{if88who}
\alias{if88who}
\title{Who would r have voted for-1988 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if88who
}
\details{
Question F. In 1988, you remember that Dukakis ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:
Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \code{1} dukakis
\item \code{2} bush
\item \code{3} other
\item \code{4} wouldnt vote
\item \code{5} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1990, 1991, 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/159/vshow}{GSS Data Explorer page for \code{if88who}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab dukakis \tab bush \tab other \tab don't know \tab no answer \tab Total \cr
   1989 \tab 1009 \tab 141 \tab 264 \tab 32 \tab 69 \tab 22 \tab 1537 \cr
   1990 \tab 913 \tab 104 \tab 252 \tab 22 \tab 72 \tab 9 \tab 1372 \cr
   1991 \tab 1017 \tab 86 \tab 292 \tab 28 \tab 74 \tab 20 \tab 1517 \cr
   1993 \tab 1045 \tab 93 \tab 282 \tab 34 \tab 112 \tab 40 \tab 1606 \cr
   Total \tab 3984 \tab 424 \tab 1090 \tab 116 \tab 327 \tab 91 \tab 6032 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
