% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{if92who}
\alias{if92who}
\title{Who would r have voted for-1992 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
if92who
}
\details{
Question G. In 1992, you remember that Clinton ran for President on the Democratic ticket against Bush for the Republicans and Perot as an Independent.
Do you remember for sure whether or not you voted in that election?
2. IF DID NOT VOTE OR INELIGIBLE:  Who would you have voted for, for President, if you had voted?
}
\section{Values}{

\itemize{
\item \verb{[1]} clinton
\item \verb{[2]} bush
\item \verb{[3]} perot
\item \verb{[4]} other
\item \verb{[5]} wouldnt vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/162/vshow}{GSS Data Explorer page for \code{if92who}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab clinton \tab bush \tab perot \tab other \tab don't know \tab no answer \tab Total \cr
   1993 \tab 1118 \tab 191 \tab 115 \tab 96 \tab 13 \tab 58 \tab 15 \tab 1606 \cr
   1994 \tab 2084 \tab 350 \tab 243 \tab 168 \tab 32 \tab 92 \tab 23 \tab 2992 \cr
   1996 \tab 1957 \tab 319 \tab 248 \tab 150 \tab 56 \tab 106 \tab 68 \tab 2904 \cr
   1998 \tab 1857 \tab 325 \tab 216 \tab 99 \tab 38 \tab 143 \tab 154 \tab 2832 \cr
   Total \tab 7016 \tab 1185 \tab 822 \tab 513 \tab 139 \tab 399 \tab 260 \tab 10334 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
