% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{immameco}
\alias{immameco}
\title{Immigrants good for america}
\source{
General Social Survey https://gss.norc.org
}
\description{
immameco
}
\details{
Question There are different opinions about immigrants from other countries living in America. (By "immigrants" we mean people who come to settle in America.) How much do you agree or disagree with each of the following statements?
B. Immigrants are generally good for America''s economy
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4860/vshow}{GSS Data Explorer page for \code{immameco}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 60 \tab 364 \tab 413 \tab 361 \tab 52 \tab 60 \tab 57 \tab 2904 \cr
   2004 \tab 1596 \tab 80 \tab 458 \tab 355 \tab 236 \tab 55 \tab 31 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 114 \tab 550 \tab 330 \tab 207 \tab 25 \tab 45 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 254 \tab 1372 \tab 1098 \tab 804 \tab 132 \tab 136 \tab 61 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immideas}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{refugees}},
\code{\link{res161}}
}
\concept{ISSP National Identity}
\keyword{variable}
