% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{immassim}
\alias{immassim}
\title{What statement about immigrants matches view}
\source{
General Social Survey https://gss.norc.org
}
\description{
immassim
}
\details{
Question Which of these statements about immigrants comes closest to your view:
}
\section{Values}{

\itemize{
\item \verb{[1]} immigrants should retain their culture of origin and not adopt american culture
\item \verb{[2]} immigrants should retain their culture of origin and also adopt american culture
\item \verb{[3]} immigrants should give up their culture of origin and adopt american culture
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4902/vshow}{GSS Data Explorer page for \code{immassim}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab immigrants should retain their culture of origin and not adopt american culture \tab immigrants should retain their culture of origin and also adopt american culture \tab immigrants should give up their culture of origin and adopt american culture \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1264 \tab 36 \tab 1084 \tab 65 \tab 85 \tab 4 \tab 2538 \cr
   Total \tab 1264 \tab 36 \tab 1084 \tab 65 \tab 85 \tab 4 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\keyword{variable}
