% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{immassim}
\alias{immassim}
\title{What statement about immigrants matches view}
\source{
General Social Survey https://gss.norc.org
}
\description{
immassim
}
\details{
Question Which of these statements about immigrants comes closest to your view:
}
\section{Values}{

\itemize{
\item \code{1} immigrants should retain their culture of origin and not adopt american culture
\item \code{2} immigrants should retain their culture of origin and also adopt american culture
\item \code{3} immigrants should give up their culture of origin and adopt american culture
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4902/vshow}{GSS Data Explorer page for \code{immassim}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\keyword{variable}
