% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{immcrime1}
\alias{immcrime1}
\title{Immigrants increase crime rates}
\source{
General Social Survey https://gss.norc.org
}
\description{
immcrime1
}
\details{
Question There are different opinions about immigrants from other countries living in America. By “immigrants” we mean people who come to settle in America. How much do you agree or disagree with each of the following statements?… Immigrants increase crime rates.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8062/vshow}{GSS Data Explorer page for \code{immcrime1}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{brneffrt}},
\code{\link{conscrp}},
\code{\link{immideas1}},
\code{\link{immjobs1}},
\code{\link{immwlfare}}
}
\concept{ISSP National Identity}
\keyword{variable}
