% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{immcrmup}
\alias{immcrmup}
\title{More immigrants cause higher crime rates}
\source{
General Social Survey https://gss.norc.org
}
\description{
immcrmup
}
\details{
Question What do you think will happen as a result of more immigrants coming to this country?  Is each of these possible results very likely, somewhat likely, not too likely, or not likely at all?
A. Higher crime rates?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not too likely
\item \code{4} not likely at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2307/vshow}{GSS Data Explorer page for \code{immcrmup}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{othhome}},
\code{\link{spklangw}},
\code{\link{uswht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
