% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{immeduc}
\alias{immeduc}
\title{Legal immigrants should have same education as americans}
\source{
General Social Survey https://gss.norc.org
}
\description{
immeduc
}
\details{
Question How much do you agree of disagree with the following statements?
C. Legal immigrants should have equal access to public education as American citizens.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4900/vshow}{GSS Data Explorer page for \code{immeduc}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\keyword{variable}
