% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{immeduc}
\alias{immeduc}
\title{Legal immigrants should have same education as americans}
\source{
General Social Survey https://gss.norc.org
}
\description{
immeduc
}
\details{
Question How much do you agree of disagree with the following statements?
C. Legal immigrants should have equal access to public education as American citizens.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4900/vshow}{GSS Data Explorer page for \code{immeduc}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab 5 \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1264 \tab 288 \tab 761 \tab 97 \tab 78 \tab 26 \tab 20 \tab 4 \tab 2538 \cr
   Total \tab 1264 \tab 288 \tab 761 \tab 97 \tab 78 \tab 26 \tab 20 \tab 4 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\keyword{variable}
