% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_173.R
\name{immfare}
\alias{immfare}
\title{Immigrants eligible for govt assistance?}
\source{
General Social Survey https://gss.norc.org
}
\description{
immfare
}
\details{
Question Under current law, immigrants who come from other countries to the United States legally are entitled, from the very beginning, to government assistance such as Medicaid, food stamps, or welfare on the same basis as citizens. But some people say they should not be eligible until they have lived here for a year or more. Which do you think? Do you think that immigrants who are here legally should be eligible for such services as soon as they come, or should they not be eligible?
}
\section{Values}{

\itemize{
\item \verb{[1]} eligible
\item \verb{[2]} not eligible
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1616/vshow}{GSS Data Explorer page for \code{immfare}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab eligible \tab not eligible \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 466 \tab 935 \tab 66 \tab 7 \tab 2992 \cr
   Total \tab 1518 \tab 466 \tab 935 \tab 66 \tab 7 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{hspasn10}},
\code{\link{immecon}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{immwrkup}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}}
}
\concept{Multiculturalism}
\keyword{variable}
