% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{immjobs}
\alias{immjobs}
\title{Immigrants take jobs away}
\source{
General Social Survey https://gss.norc.org
}
\description{
immjobs
}
\details{
Question There are different opinions about immigrants from other countries living in America. (By "immigrants" we mean people who come to settle in America.) How much do you agree or disagree with each of the following statements?
C. Immigrants take jobs away from people who were born in America
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   A/C/- \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4861/vshow}{GSS Data Explorer page for \code{immjobs}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1996 \tab 1537 \tab 170 \tab 446 \tab 299 \tab 301 \tab 65 \tab 28 \tab 58 \tab - \tab 2904 \cr
   2004 \tab 1596 \tab 153 \tab 363 \tab 274 \tab 343 \tab 67 \tab 15 \tab 1 \tab - \tab 2812 \cr
   2014 \tab 1264 \tab 94 \tab 349 \tab 271 \tab 438 \tab 87 \tab 31 \tab 4 \tab - \tab 2538 \cr
   2022 \tab 1203 \tab 147 \tab 375 \tab 505 \tab 804 \tab 487 \tab 17 \tab 3 \tab 3 \tab 3544 \cr
   Total \tab 5600 \tab 564 \tab 1533 \tab 1349 \tab 1886 \tab 706 \tab 91 \tab 66 \tab 3 \tab 11798 \cr
}
}

\seealso{
Other Core: 
\code{\link{immcrime}}
}
\concept{Core}
\keyword{variable}
