% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{immwlfare}
\alias{immwlfare}
\title{People born in america should be given preference over immigrants when it comes to jobs, housing, or health care}
\source{
General Social Survey https://gss.norc.org
}
\description{
immwlfare
}
\details{
Question People born in America should be given preference over immigrants when it comes to jobs, housing, or health care.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8065/vshow}{GSS Data Explorer page for \code{immwlfare}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2024 \tab - \tab 207 \tab 370 \tab 373 \tab 426 \tab 171 \tab 82 \tab 1663 \tab 2 \tab 15 \tab 3309 \cr
   Total \tab 0 \tab 207 \tab 370 \tab 373 \tab 426 \tab 171 \tab 82 \tab 1663 \tab 2 \tab 15 \tab 3309 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{bizmoney}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{experts}},
\code{\link{fulldem}},
\code{\link{incgap1}}
}
\concept{ISSP National Identity}
\keyword{variable}
