% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_173.R
\name{immwrkup}
\alias{immwrkup}
\title{Should immigrants overcome bias without help}
\source{
General Social Survey https://gss.norc.org
}
\description{
immwrkup
}
\details{
Question Now I'm going to read you some statements and would like to get your reaction to them. After I read each statement, please tell me if you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree with the statement.
B. The Irish, Italians, Jews, and many other minorities overcame prejudice and worked their way up. Today's immigrants should do the same without any special favors.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1621/vshow}{GSS Data Explorer page for \code{immwrkup}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 500 \tab 673 \tab 155 \tab 84 \tab 15 \tab 36 \tab 11 \tab 2992 \cr
   Total \tab 1518 \tab 500 \tab 673 \tab 155 \tab 84 \tab 15 \tab 36 \tab 11 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{hspasn10}},
\code{\link{immecon}},
\code{\link{immfare}},
\code{\link{immpush}},
\code{\link{immunemp}},
\code{\link{immunite}},
\code{\link{undoccol}},
\code{\link{undockid}},
\code{\link{undocwrk}}
}
\concept{Multiculturalism}
\keyword{variable}
