% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_85.R
\name{impfam}
\alias{impfam}
\title{Family and children}
\source{
General Social Survey https://gss.norc.org
}
\description{
impfam
}
\details{
Question On these cards are various aspects of life. We would like to know how important each of these aspects of life is for you. On each of these Cards you see on the right hand side a scale with seven points. The lowest point with number 1 indicates that this aspect of life is unimportant to you. Point 7 at the top indicates that particular aspect of life is very important to you.
The numbers in between indicate varying degrees of importance. Please take a look at all the cards first. Then, tell me for each card its letter and the number you've decided on. SHUFFLE THE SET OF BLUE CardS BEFORE HANDING TO RESPONDENT.
A. One's own family and children.
}
\section{Values}{

\itemize{
\item \verb{[1]} unimportant
\item \verb{[7]} very important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/765/vshow}{GSS Data Explorer page for \code{impfam}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab unimportant \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 26 \tab 8 \tab 13 \tab 12 \tab 48 \tab 82 \tab 1647 \tab 2 \tab 22 \tab 1860 \cr
   Total \tab 0 \tab 26 \tab 8 \tab 13 \tab 12 \tab 48 \tab 82 \tab 1647 \tab 2 \tab 22 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{cutspdfg}},
\code{\link{cutspdr}},
\code{\link{hlpminfg}},
\code{\link{hlpminr}},
\code{\link{impchurh}},
\code{\link{impfrend}},
\code{\link{impkin}},
\code{\link{imprelax}},
\code{\link{impwork}}
}
\concept{Core}
\keyword{variable}
