% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_152.R
\name{impfinan}
\alias{impfinan}
\title{Imp to r of financial security}
\source{
General Social Survey https://gss.norc.org
}
\description{
impfinan
}
\details{
Question I'm going to read you a list of some things that different people value. Some people say these things are very important to them. Other people say they are not so important. Please tell me how important each things is to you personally, using the responses on this card (HAND CARD TO RESPONDENT). How about being financially secure? Is it one of the most important values you hold, very important, somewhat important, not too important, or not at all important?.
A. Being financially secure.
}
\section{Values}{

\itemize{
\item \verb{[1]} one of most important
\item \verb{[2]} very important
\item \verb{[3]} somewhat important
\item \verb{[4]} not too important
\item \verb{[5]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1438/vshow}{GSS Data Explorer page for \code{impfinan}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab one of most important \tab very important \tab somewhat important \tab not too important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 427 \tab 817 \tab 290 \tab 49 \tab 6 \tab 5 \tab 12 \tab 1606 \cr
   Total \tab 0 \tab 427 \tab 817 \tab 290 \tab 49 \tab 6 \tab 5 \tab 12 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{excelart}},
\code{\link{frdcreat}},
\code{\link{frdcultr}},
\code{\link{frddynam}},
\code{\link{frdfun}},
\code{\link{frdhonst}},
\code{\link{frdintel}},
\code{\link{frdresp}},
\code{\link{hosthome}}
}
\concept{Culture}
\keyword{variable}
