% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_153.R
\name{impjob}
\alias{impjob}
\title{Imp to r of having a fulfilling job}
\source{
General Social Survey https://gss.norc.org
}
\description{
impjob
}
\details{
Question I'm going to read you a list of some things that different people value. Some people say these things are very important to them. Other people say they are not so important. Please tell me how important each things is to you personally, using the responses on this card (HAND CARD TO RESPONDENT). How about being financially secure? Is it one of the most important values you hold, very important, somewhat important, not too important, or not at all important?
G. Having a fulfilling job.
}
\section{Values}{

\itemize{
\item \code{1} one of most important
\item \code{2} very important
\item \code{3} somewhat important
\item \code{4} not too important
\item \code{5} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1444/vshow}{GSS Data Explorer page for \code{impjob}}.
}

\seealso{
Other Culture: 
\code{\link{impcultr}},
\code{\link{impgod}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}},
\code{\link{mostimp1}},
\code{\link{mostimp2}},
\code{\link{mostimp3}}

Other Values: 
\code{\link{impcultr}},
\code{\link{impgod}},
\code{\link{impkids}},
\code{\link{impmar}},
\code{\link{impself}},
\code{\link{impthngs}}
}
\concept{Culture}
\concept{Values}
\concept{Work}
\keyword{variable}
