% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{imppromo}
\alias{imppromo}
\title{Importance of the opportunity to be promoted}
\source{
General Social Survey https://gss.norc.org
}
\description{
imppromo
}
\details{
Question How important to you is the opportunity to be promoted?  Is it…
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} somewhat important
\item \verb{[4]} not at all important
\item \verb{[5]} does not apply, volunteered
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1302/vshow}{GSS Data Explorer page for \code{imppromo}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab somewhat important \tab not at all important \tab does not apply, volunteered \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 296 \tab 181 \tab 142 \tab 165 \tab 115 \tab 1 \tab 12 \tab 1517 \cr
   Total \tab 605 \tab 296 \tab 181 \tab 142 \tab 165 \tab 115 \tab 1 \tab 12 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{futpromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}
}
\concept{Work Organizations}
\keyword{variable}
