% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{imprvtrt}
\alias{imprvtrt}
\title{How likely x s situation will improve with treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
imprvtrt
}
\details{
Question In your opinion, how likely is it that NAME's situation will improve with treatment -- very likely, somewhat likely, somewhat unlikely, or not likely at all?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} not likely at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2002/vshow}{GSS Data Explorer page for \code{imprvtrt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab somewhat unlikely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 690 \tab 590 \tab 53 \tab 50 \tab 42 \tab 19 \tab 2904 \cr
   2006 \tab 3073 \tab 673 \tab 621 \tab 53 \tab 21 \tab 55 \tab 14 \tab 4510 \cr
   2018 \tab 1175 \tab 577 \tab 468 \tab 51 \tab 29 \tab 20 \tab 28 \tab 2348 \cr
   Total \tab 5708 \tab 1940 \tab 1679 \tab 157 \tab 100 \tab 117 \tab 61 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
