% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{imprvtrt}
\alias{imprvtrt}
\title{How likely x s situation will improve with treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
imprvtrt
}
\details{
Question In your opinion, how likely is it that NAME's situation will improve with treatment -- very likely, somewhat likely, somewhat unlikely, or not likely at all?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} not likely at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2002/vshow}{GSS Data Explorer page for \code{imprvtrt}}.
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}

Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
