% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_472.R
\name{incgap}
\alias{incgap}
\title{Income differentials in usa too big}
\source{
General Social Survey https://gss.norc.org
}
\description{
incgap
}
\details{
Question Do you agree or disagree?
A. Differences in income in America are too large.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4295/vshow}{GSS Data Explorer page for \code{incgap}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givrocc}},
\code{\link{govedop}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}}

Other Inequality, Income: 
\code{\link{givfctry}},
\code{\link{givjudge}},
\code{\link{givrocc}},
\code{\link{goveqinc}},
\code{\link{govjobs}},
\code{\link{govless}},
\code{\link{govminc}},
\code{\link{govunemp}}
}
\concept{ISSP Social Inequality}
\concept{Inequality, Income}
\concept{Markets}
\keyword{variable}
