% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{inclstjb}
\alias{inclstjb}
\title{Earnings from all sources for this job}
\source{
General Social Survey https://gss.norc.org
}
\description{
inclstjb
}
\details{
Question Thinking about your last job, in which of these groups did your earnings from all sources for this job fall? that is, before taxes or other deductions.
}
\section{Values}{

\itemize{
\item \code{1} a. under $1,000
\item \code{2} b. $1,000 to 2,999
\item \code{3} c. $3,000 to 3,999
\item \code{4} d. $4,000 to 4,999
\item \code{5} e. $5,000 to 5,999
\item \code{6} f. $6,000 to 6,999
\item \code{7} g. $7,000 to 7,999
\item \code{8} h. $8,000 to 9,999
\item \code{9} i. $10,000 to 12,499
\item \code{10} j. $12,500 to 14,999
\item \code{11} k. $15,000 to 17,499
\item \code{12} l. $17,500 to 19,999
\item \code{13} m. $20,000 to 22,499
\item \code{14} n. $22,500 to 24,999
\item \code{15} o. $25,000 to 29,999
\item \code{16} p. $30,000 to 34,999
\item \code{17} q. $35,000 to 39,999
\item \code{18} r. $40,000 to 49,999
\item \code{19} s. $50,000 to 59,999
\item \code{20} t. $60,000 to 74,999
\item \code{21} u. $75,000 to $89,999
\item \code{22} v. $90,000 to $109,999
\item \code{23} w. $110,000 to $129,999
\item \code{24} x. $130,000 to $149,999
\item \code{25} y. $150,000 or over
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3897/vshow}{GSS Data Explorer page for \code{inclstjb}}.
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
