% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{inclstjb}
\alias{inclstjb}
\title{Earnings from all sources for this job}
\source{
General Social Survey https://gss.norc.org
}
\description{
inclstjb
}
\details{
Question Thinking about your last job, in which of these groups did your earnings from all sources for this job fall? that is, before taxes or other deductions.
}
\section{Values}{

\itemize{
\item \verb{[1]} a. under $1,000
\item \verb{[2]} b. $1,000 to 2,999
\item \verb{[3]} c. $3,000 to 3,999
\item \verb{[4]} d. $4,000 to 4,999
\item \verb{[5]} e. $5,000 to 5,999
\item \verb{[6]} f. $6,000 to 6,999
\item \verb{[7]} g. $7,000 to 7,999
\item \verb{[8]} h. $8,000 to 9,999
\item \verb{[9]} i. $10,000 to 12,499
\item \verb{[10]} j. $12,500 to 14,999
\item \verb{[11]} k. $15,000 to 17,499
\item \verb{[12]} l. $17,500 to 19,999
\item \verb{[13]} m. $20,000 to 22,499
\item \verb{[14]} n. $22,500 to 24,999
\item \verb{[15]} o. $25,000 to 29,999
\item \verb{[16]} p. $30,000 to 34,999
\item \verb{[17]} q. $35,000 to 39,999
\item \verb{[18]} r. $40,000 to 49,999
\item \verb{[19]} s. $50,000 to 59,999
\item \verb{[20]} t. $60,000 to 74,999
\item \verb{[21]} u. $75,000 to $89,999
\item \verb{[22]} v. $90,000 to $109,999
\item \verb{[23]} w. $110,000 to $129,999
\item \verb{[24]} x. $130,000 to $149,999
\item \verb{[25]} y. $150,000 or over
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3897/vshow}{GSS Data Explorer page for \code{inclstjb}}.

Counts by year:\tabular{llllllllllllllllllllllllllllll}{
   year \tab iap \tab a. under $1,000 \tab b. $1,000 to 2,999 \tab c. $3,000 to 3,999 \tab d. $4,000 to 4,999 \tab e. $5,000 to 5,999 \tab f. $6,000 to 6,999 \tab g. $7,000 to 7,999 \tab h. $8,000 to 9,999 \tab i. $10,000 to 12,499 \tab j. $12,500 to 14,999 \tab k. $15,000 to 17,499 \tab l. $17,500 to 19,999 \tab m. $20,000 to 22,499 \tab n. $22,500 to 24,999 \tab o. $25,000 to 29,999 \tab p. $30,000 to 34,999 \tab q. $35,000 to 39,999 \tab r. $40,000 to 49,999 \tab s. $50,000 to 59,999 \tab t. $60,000 to 74,999 \tab u. $75,000 to $89,999 \tab v. $90,000 to $109,999 \tab w. $110,000 to $129,999 \tab x. $130,000 to $149,999 \tab y. $150,000 or over \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1822 \tab 5 \tab 5 \tab 6 \tab 4 \tab 7 \tab 2 \tab 6 \tab 11 \tab 14 \tab 10 \tab 9 \tab 8 \tab 10 \tab 9 \tab 18 \tab 11 \tab 9 \tab 15 \tab 6 \tab 6 \tab 6 \tab 4 \tab 2 \tab 2 \tab 1 \tab 7 \tab 8 \tab 2023 \cr
   Total \tab 1822 \tab 5 \tab 5 \tab 6 \tab 4 \tab 7 \tab 2 \tab 6 \tab 11 \tab 14 \tab 10 \tab 9 \tab 8 \tab 10 \tab 9 \tab 18 \tab 11 \tab 9 \tab 15 \tab 6 \tab 6 \tab 6 \tab 4 \tab 2 \tab 2 \tab 1 \tab 7 \tab 8 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
