% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_282.R
\name{incminw}
\alias{incminw}
\title{Minutes per week using work computer make money}
\source{
General Social Survey https://gss.norc.org
}
\description{
incminw
}
\details{
Question For about how many minutes or hours a week do you use your work computer for each of the following:
B. Other income-producing activities?
\enumerate{
\item Minutes per week?
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2569/vshow}{GSS Data Explorer page for \code{incminw}}.
}

\seealso{
Other Information Society: 
\code{\link{ecomhrw}},
\code{\link{ecomminw}},
\code{\link{emhrw}},
\code{\link{emminw}},
\code{\link{inchrw}},
\code{\link{perhrw}},
\code{\link{perminw}},
\code{\link{wwwhrw}},
\code{\link{wwwminw}}
}
\concept{Information Society}
\keyword{variable}
