% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_64.R
\name{incneed}
\alias{incneed}
\title{Income family needs}
\source{
General Social Survey https://gss.norc.org
}
\description{
incneed
}
\details{
Question Living where you do now and meeting the expenses you consider necessary, what would be the very smallest amount of income per month--after taxes--your household would need to make ends meet?
}
\section{Values}{

\itemize{
\item \code{99997} depends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/575/vshow}{GSS Data Explorer page for \code{incneed}}.
}

\seealso{
Other Core: 
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{mininc}},
\code{\link{satfin}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}}
}
\concept{Core}
\keyword{variable}
