% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{income06}
\alias{income06}
\title{Total family income}
\source{
General Social Survey https://gss.norc.org
}
\description{
income06
}
\details{
Question In which of these groups did your total family income, from all sources, fall last year before taxes, that is.
Just tell me the letter.
}
\section{Values}{

\itemize{
\item \code{1} under $1 000
\item \code{2} $1 000 to 2 999
\item \code{3} $3 000 to 3 999
\item \code{4} $4 000 to 4 999
\item \code{5} $5 000 to 5 999
\item \code{6} $6 000 to 6 999
\item \code{7} $7 000 to 7 999
\item \code{8} $8 000 to 9 999
\item \code{9} $10000 to 12499
\item \code{10} $12500 to 14999
\item \code{11} $15000 to 17499
\item \code{12} $17500 to 19999
\item \code{13} $20000 to 22499
\item \code{14} $22500 to 24999
\item \code{15} $25000 to 29999
\item \code{16} $30000 to 34999
\item \code{17} $35000 to 39999
\item \code{18} $40000 to 49999
\item \code{19} $50000 to 59999
\item \code{20} $60000 to 74999
\item \code{21} $75000 to $89999
\item \code{22} $90000 to $109999
\item \code{23} $110000 to $129999
\item \code{24} $130000 to $149999
\item \code{25} $150000 or over
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/117/vshow}{GSS Data Explorer page for \code{income06}}.
}

\seealso{
Other Core: 
\code{\link{dotdata}},
\code{\link{income16}},
\code{\link{region}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom98}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}

Other Income: 
\code{\link{rincom06}},
\code{\link{rincom98}}
}
\concept{Core}
\concept{Income}
\keyword{variable}
