% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{income16}
\alias{income16}
\title{Total family income}
\source{
General Social Survey https://gss.norc.org
}
\description{
income16
}
\details{
Question In which of these groups did your total family income, from all sources, fall last year -- 2015 -- before taxes, that is. Just tell me the letter.                       
Total income includes interest or dividends, rent, Social Security, other pensions,  alimony or child support, unemployment compensation, public aid (welfare),  armed forces or veteran's allotment.
}
\section{Values}{

\itemize{
\item \code{1} under $1,000
\item \code{2} $1,000 to $2,999
\item \code{3} $3,000 to $3,999
\item \code{4} $4,000 to $4,999
\item \code{5} $5,000 to $5,999
\item \code{6} $6,000 to $6,999
\item \code{7} $7,000 to $7,999
\item \code{8} $8,000 to $9,999
\item \code{9} $10,000 to $12,499
\item \code{10} $12,500 to $14,999
\item \code{11} $15,000 to $17,499
\item \code{12} $17,500 to $19,999
\item \code{13} $20,000 to $22,499
\item \code{14} $22,500 to $24,999
\item \code{15} $25,000 to $29,999
\item \code{16} $30,000 to $34,999
\item \code{17} $35,000 to $39,999
\item \code{18} $40,000 to $49,999
\item \code{19} $50,000 to $59,999
\item \code{20} $60,000 to $74,999
\item \code{21} $75,000 to $89,999
\item \code{22} $90,000 to $109,999
\item \code{23} $110,000 to $129,999
\item \code{24} $130,000 to $149,999
\item \code{25} $150,000 to $169,999
\item \code{26} $170,000 or over
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6167/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab $1,000 to $2,999 \tab $10,000 to $12,499 \tab $110,000 to $129,999 \tab $12,500 to $14,999 \tab $130,000 to $149,999 \tab $15,000 to $17,499 \tab $150,000 to $169,999 \tab $17,500 to $19,999 \tab $170,000 or over \tab $20,000 to $22,499 \tab $22,500 to $24,999 \tab $25,000 to $29,999 \tab $3,000 to $3,999 \tab $30,000 to $34,999 \tab $35,000 to $39,999 \tab $4,000 to $4,999 \tab $40,000 to $49,999 \tab $5,000 to $5,999 \tab $50,000 to $59,999 \tab $6,000 to $6,999 \tab $60,000 to $74,999 \tab $7,000 to $7,999 \tab $75,000 to $89,999 \tab $8,000 to $9,999 \tab $90,000 to $109,999 \tab don't know \tab refused \tab under $1,000 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 35 \tab 88 \tab 126 \tab 71 \tab 99 \tab 65 \tab 61 \tab 65 \tab 164 \tab 95 \tab 101 \tab 115 \tab 24 \tab 131 \tab 126 \tab 13 \tab 218 \tab 19 \tab 207 \tab 15 \tab 258 \tab 23 \tab 214 \tab 48 \tab 179 \tab 134 \tab 137 \tab 36 \tab - \tab 2867 \cr
   2018 \tab - \tab 27 \tab 70 \tab 128 \tab 70 \tab 77 \tab 44 \tab 72 \tab 37 \tab 171 \tab 65 \tab 77 \tab 80 \tab 19 \tab 116 \tab 101 \tab 14 \tab 159 \tab 15 \tab 187 \tab 9 \tab 215 \tab 13 \tab 167 \tab 41 \tab 145 \tab 91 \tab 105 \tab 33 \tab - \tab 2348 \cr
   2021 \tab - \tab 50 \tab 87 \tab 207 \tab 80 \tab 141 \tab 74 \tab 140 \tab 66 \tab 372 \tab 80 \tab 69 \tab 131 \tab 19 \tab 155 \tab 136 \tab 11 \tab 266 \tab 17 \tab 253 \tab 12 \tab 367 \tab 18 \tab 314 \tab 54 \tab 326 \tab 335 \tab 80 \tab 64 \tab 108 \tab 4032 \cr
   2022 \tab - \tab 33 \tab 104 \tab 169 \tab 83 \tab 147 \tab 64 \tab 97 \tab 51 \tab 330 \tab 92 \tab 100 \tab 129 \tab 18 \tab 162 \tab 152 \tab 11 \tab 211 \tab 16 \tab 239 \tab 14 \tab 304 \tab 16 \tab 240 \tab 40 \tab 249 \tab 268 \tab 132 \tab 39 \tab 34 \tab 3544 \cr
   2024 \tab - \tab 48 \tab 97 \tab 165 \tab 78 \tab 130 \tab 51 \tab 107 \tab 52 \tab 355 \tab 77 \tab 62 \tab 91 \tab 26 \tab 123 \tab 126 \tab 18 \tab 185 \tab 25 \tab 236 \tab 14 \tab 268 \tab 16 \tab 235 \tab 34 \tab 245 \tab 246 \tab 77 \tab 67 \tab 55 \tab 3309 \cr
   Total \tab 59599 \tab 193 \tab 446 \tab 795 \tab 382 \tab 594 \tab 298 \tab 477 \tab 271 \tab 1392 \tab 409 \tab 409 \tab 546 \tab 106 \tab 687 \tab 641 \tab 67 \tab 1039 \tab 92 \tab 1122 \tab 64 \tab 1412 \tab 86 \tab 1170 \tab 217 \tab 1144 \tab 1074 \tab 531 \tab 239 \tab 197 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\seealso{
Other Core: 
\code{\link{dotdata}},
\code{\link{income06}},
\code{\link{rincom06}},
\code{\link{rincom16}},
\code{\link{rincom98}},
\code{\link{size}},
\code{\link{srcbelt}},
\code{\link{xnorcsiz}}
}
\concept{Core}
\keyword{variable}
