% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_16.R
\name{income72}
\alias{income72}
\title{Total family income}
\source{
General Social Survey https://gss.norc.org
}
\description{
income72
}
\details{
Question In which of these groups did your total family income, from all sources, fall last year--1971--before taxes, that is?
}
\section{Values}{

\itemize{
\item \code{1} lt $2000
\item \code{2} $2000 to $3999
\item \code{3} $4000 to $5999
\item \code{4} $6000 to $7999
\item \code{5} $8000 to $9999
\item \code{6} $10000 to $12499
\item \code{7} $12500 to $14999
\item \code{8} $15000 to $17499
\item \code{9} $17500 to $19999
\item \code{10} $20000 to $24999
\item \code{11} $25000 to $29999
\item \code{12} $30000+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/106/vshow}{GSS Data Explorer page for \code{income72}}.
}

\seealso{
Other Core: 
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}}

Other Income: 
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{rincom77}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}}
}
\concept{Core}
\concept{Income}
\keyword{variable}
