% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{incuspop}
\alias{incuspop}
\title{Estimated income status of housing unit}
\source{
General Social Survey https://gss.norc.org
}
\description{
incuspop
}
\details{
Question What do you believe the income status of this housing unit is, relative to the rest of the general United States population?
}
\section{Values}{

\itemize{
\item \code{1} higher than average
\item \code{2} average
\item \code{3} lower than average
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6657/vshow}{GSS Data Explorer page for \code{incuspop}}.
}

\keyword{variable}
