% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{indusgen}
\alias{indusgen}
\title{Indust air pollution danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
indusgen
}
\details{
Question A. In general, do you think that air pollution caused by industry is…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{chemgen}},
\code{\link{tempgen}},
\code{\link{watergen}}

Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
