% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{indusgen1}
\alias{indusgen1}
\title{In general, do you think that air pollution caused by industry is}
\source{
General Social Survey https://gss.norc.org
}
\description{
indusgen1
}
\details{
Question In general, do you think that air pollution caused by industry is…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7490/vshow}{GSS Data Explorer page for \code{indusgen1}}.
}

\keyword{variable}
