% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{indusoth}
\alias{indusoth}
\title{R preference re: manufacture service, etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
indusoth
}
\details{
Question Suppose you were working and could choose between different kinds of jobs. Which of the following would you personally choose?  I would advise...
Please CIRCLE ONE CODE FOR EACH.
C. Working in a manufacturing industry (or) in an office, sales, or service
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacture
\item \verb{[2]} office, sales, etc
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{iwrkhard}},
\code{\link{privgovt}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
