% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{ineqmad}
\alias{ineqmad}
\title{Feeling about differences in wealth}
\source{
General Social Survey https://gss.norc.org
}
\description{
ineqmad
}
\details{
Question How do you feel when you think about differences in wealth between the rich and the poor in America?
}
\section{Values}{

\itemize{
\item \code{0} 0 - not angry at all
\item \code{10} 10 - extremely angry
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7509/vshow}{GSS Data Explorer page for \code{ineqmad}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 0 - not angry at all \tab 1 \tab 10 - extremely angry \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 266 \tab 76 \tab 132 \tab 109 \tab 124 \tab 125 \tab 356 \tab 168 \tab 229 \tab 143 \tab 50 \tab 68 \tab 147 \tab 10 \tab 4032 \cr
   Total \tab 2029 \tab 266 \tab 76 \tab 132 \tab 109 \tab 124 \tab 125 \tab 356 \tab 168 \tab 229 \tab 143 \tab 50 \tab 68 \tab 147 \tab 10 \tab 4032 \cr
}
}

\keyword{variable}
