% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{inequal1}
\alias{inequal1}
\title{Extra pay -) people take responsibility}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal1
}
\details{
Question Do you agree or disagree . . .
A. People would not want to take extra responsibility at work unless they were paid extra for it.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4256/vshow}{GSS Data Explorer page for \code{inequal1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 297 \tab 788 \tab 170 \tab 234 \tab 30 \tab 17 \tab 28 \tab 1819 \cr
   1996 \tab 1460 \tab 526 \tab 592 \tab 74 \tab 174 \tab 56 \tab 14 \tab 8 \tab 2904 \cr
   Total \tab 1715 \tab 823 \tab 1380 \tab 244 \tab 408 \tab 86 \tab 31 \tab 36 \tab 4723 \cr
}
}

\seealso{
Other Markets: 
\code{\link{inequal3}},
\code{\link{inequal5}}

Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{incentiv}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{oppol}},
\code{\link{opsex}}
}
\concept{ISSP Social Inequality}
\concept{Markets}
\keyword{variable}
