% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{inequal2}
\alias{inequal2}
\title{Extra pay -) workers acquire skills}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal2
}
\details{
Question Do you agree or disagree . . .
B. Workers would not bother to get skills and qualifications unless they were paid extra for having them.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4257/vshow}{GSS Data Explorer page for \code{inequal2}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{incentiv}},
\code{\link{inequal1}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{oppol}},
\code{\link{opsex}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
