% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{inequal3}
\alias{inequal3}
\title{Inequality exists for benefit of rich}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal3
}
\details{
Question Do you agree or disagree . . .
C. Inequality continues to exist because it benefits the rich and powerful.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4258/vshow}{GSS Data Explorer page for \code{inequal3}}.
}

\seealso{
Other Markets: 
\code{\link{inequal1}},
\code{\link{inequal5}}

Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{incentiv}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{oppol}},
\code{\link{opsex}}

Other Inequality, Income: 
\code{\link{incentiv}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal4}},
\code{\link{inequal5}}
}
\concept{ISSP Social Inequality}
\concept{Inequality, Income}
\concept{Markets}
\keyword{variable}
