% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{inequal4}
\alias{inequal4}
\title{Extra pay -) people earn degrees}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal4
}
\details{
Question Do you agree or disagree . . .
D. No one would study for years to become a lawyer or doctor unless they expected to earn a lot more than ordinary workers.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4259/vshow}{GSS Data Explorer page for \code{inequal4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 419 \tab 643 \tab 165 \tab 260 \tab 32 \tab 12 \tab 33 \tab 1819 \cr
   2000 \tab 1545 \tab 264 \tab 472 \tab 171 \tab 235 \tab 45 \tab 12 \tab 73 \tab 2817 \cr
   Total \tab 1800 \tab 683 \tab 1115 \tab 336 \tab 495 \tab 77 \tab 24 \tab 106 \tab 4636 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{incentiv}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal5}},
\code{\link{oppol}},
\code{\link{opsex}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
