% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{inequal5}
\alias{inequal5}
\title{Pay differences -) american prosperity}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal5
}
\details{
Question Do you agree or disagree . . .
E. Large differences in income are necessary for America''s prosperity.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2000 \cr
   B/C/- \tab 2008, 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4260/vshow}{GSS Data Explorer page for \code{inequal5}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 102 \tab 394 \tab 419 \tab 458 \tab 85 \tab 66 \tab 40 \tab - \tab 1819 \cr
   1996 \tab 1460 \tab 114 \tab 287 \tab 177 \tab 434 \tab 361 \tab 64 \tab 7 \tab - \tab 2904 \cr
   2000 \tab 1545 \tab 57 \tab 260 \tab 347 \tab 394 \tab 101 \tab 38 \tab 75 \tab - \tab 2817 \cr
   2008 \tab 982 \tab 28 \tab 216 \tab 177 \tab 450 \tab 134 \tab 29 \tab 7 \tab - \tab 2023 \cr
   2021 \tab 2252 \tab 35 \tab 161 \tab 451 \tab 664 \tab 373 \tab 75 \tab 3 \tab 18 \tab 4032 \cr
   Total \tab 6494 \tab 336 \tab 1318 \tab 1571 \tab 2400 \tab 1054 \tab 272 \tab 132 \tab 18 \tab 13595 \cr
}
}

\seealso{
Other Markets: 
\code{\link{inequal1}},
\code{\link{inequal3}}

Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{incentiv}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{oppol}},
\code{\link{opsex}}
}
\concept{ISSP Social Inequality}
\concept{Markets}
\keyword{variable}
