% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{inequal6}
\alias{inequal6}
\title{Everyone benefits from business profits}
\source{
General Social Survey https://gss.norc.org
}
\description{
inequal6
}
\details{
Question Do you agree or disagree . . .
F. Allowing business to make good profits is the best way to improve everyone''s standard of living.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7551/vshow}{GSS Data Explorer page for \code{inequal6}}.
}

\keyword{variable}
