% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{infodeal}
\alias{infodeal}
\title{Would r provide info for discounts, free products}
\source{
General Social Survey https://gss.norc.org
}
\description{
infodeal
}
\details{
Question Private companies offer services in return for personal information such as your name, phone number, email address, online-search history, or purchases. Regardless of whether you use the Internet yourself, please indicate the extent to which you agree or disagree with the following statements.… I would provide my personal information online in return for discounts and free products.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8052/vshow}{GSS Data Explorer page for \code{infodeal}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{classtech}},
\code{\link{dataprot}},
\code{\link{eductech}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infoprofit}},
\code{\link{trcong}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
