% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{infpubtv}
\alias{infpubtv}
\title{Influence of children's programs on public tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
infpubtv
}
\details{
Question Children are exposed to many influences in their daily lives. What kind of influence does each of the following have on children: Very positive, positive, neutral, negative, very negative?:
B. Children''s programs on television.
}
\section{Values}{

\itemize{
\item \verb{[1]} very positive
\item \verb{[2]} positive
\item \verb{[3]} neutral
\item \verb{[4]} negative
\item \verb{[5]} very negative
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5118/vshow}{GSS Data Explorer page for \code{infpubtv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very positive \tab positive \tab neutral \tab negative \tab very negative \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 299 \tab 811 \tab 146 \tab 54 \tab 17 \tab 38 \tab 7 \tab 1372 \cr
   Total \tab 0 \tab 299 \tab 811 \tab 146 \tab 54 \tab 17 \tab 38 \tab 7 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{sppregnt}}
}
\concept{Children}
\keyword{variable}
