% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_346.R
\name{infrmgrp}
\alias{infrmgrp}
\title{Any activities with the same group of ppl regularly}
\source{
General Social Survey https://gss.norc.org
}
\description{
infrmgrp
}
\details{
Question Are there any activities that you do with the same group of people on a regular basis even if the group doesn’t have a name such as a bridge group, exercise group, or a group that meets to discuss individual or community problems?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3109/vshow}{GSS Data Explorer page for \code{infrmgrp}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{churchtx}},
\code{\link{typfrat}},
\code{\link{typpolit}},
\code{\link{typserv}},
\code{\link{typvet}}

Other Voluntary Associations: 
\code{\link{churchtx}},
\code{\link{typfrat}},
\code{\link{typpolit}},
\code{\link{typserv}},
\code{\link{typvet}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
