% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{inlawnum}
\alias{inlawnum}
\title{Number of in-laws}
\source{
General Social Survey https://gss.norc.org
}
\description{
inlawnum
}
\details{
Question Now thinking of all your adult relatives -- those still living and age 18 or older.
How many of each do you have?  READ EACH ITEM. CODE A NUMBER OR ZERO FOR EACH.
D. Parents-in-Law and Adult Brothers- and Sisters-in-Law
}
\section{Values}{

\itemize{
\item \code{96} 96 or more
\item \code{97} # dk but ) 0
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4156/vshow}{GSS Data Explorer page for \code{inlawnum}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}

Other Family: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}
}
\concept{Family}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
