% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{inperson}
\alias{inperson}
\title{How many people seen in person}
\source{
General Social Survey https://gss.norc.org
}
\description{
inperson
}
\details{
Question Of these____friends and relatives, about how many do you stay in contact with by:
A. Seeing them socially, face-to-face?
}
\section{Values}{

\itemize{
\item \verb{[1]} 0
\item \verb{[2]} 1 - 2
\item \verb{[3]} 3 - 5
\item \verb{[4]} 6 - 10
\item \verb{[5]} 11 - 15
\item \verb{[6]} 16-25
\item \verb{[7]} 26-50
\item \verb{[8]} 50+
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2544/vshow}{GSS Data Explorer page for \code{inperson}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 - 2 \tab 3 - 5 \tab 6 - 10 \tab 11 - 15 \tab 16-25 \tab 26-50 \tab 50+ \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1884 \tab 63 \tab 139 \tab 213 \tab 192 \tab 83 \tab 96 \tab 76 \tab 44 \tab 4 \tab 23 \tab 2817 \cr
   2002 \tab 74 \tab 103 \tab 322 \tab 531 \tab 572 \tab 302 \tab 361 \tab 313 \tab 173 \tab 1 \tab 13 \tab 2765 \cr
   2004 \tab 1873 \tab 66 \tab 133 \tab 183 \tab 197 \tab 102 \tab 113 \tab 79 \tab 50 \tab 4 \tab 12 \tab 2812 \cr
   Total \tab 3831 \tab 232 \tab 594 \tab 927 \tab 961 \tab 487 \tab 570 \tab 468 \tab 267 \tab 9 \tab 48 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{letters}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}},
\code{\link{polinfgn}}
}
\concept{Information Society}
\keyword{variable}
