% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{instype02}
\alias{instype02}
\title{Second type of insurance mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
instype02
}
\details{
Question What kind of health insurance do you have?  Please check all that apply.  Second type of insurance mentioned.
}
\section{Values}{

\itemize{
\item \code{1} public health insurance
\item \code{2} private insurance
\item \code{3} employer/union based insurance
\item \code{4} other
\item \code{5} have no health insurance
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7973/vshow}{GSS Data Explorer page for \code{instype02}}.
}

\keyword{variable}
