% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_673.R
\name{intaccprxy}
\alias{intaccprxy}
\title{How often r asked someone to use the internet on their behalf}
\source{
General Social Survey https://gss.norc.org
}
\description{
intaccprxy
}
\details{
Question In the last 12 months how often, if at all, did you ask someone to do something for you on the Internet on your behalf, such as sending an email, getting information or shopping online?
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} once or twice
\item \code{3} more often
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8114/vshow}{GSS Data Explorer page for \code{intaccprxy}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{ftruse}},
\code{\link{inthome}},
\code{\link{intprfssnl}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{prxyuse}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
