% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intapps}
\alias{intapps}
\title{R's skill at learning new software}
\source{
General Social Survey https://gss.norc.org
}
\description{
intapps
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} very good
\item \code{2} good
\item \code{3} neither good nor bad
\item \code{4} bad
\item \code{5} very bad
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
