% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{intbhlf}
\alias{intbhlf}
\title{Frequency r was asked to use internet for someone else}
\source{
General Social Survey https://gss.norc.org
}
\description{
intbhlf
}
\details{
Question In the last 12 months, how often, if at all, were you asked by someone to do something on their behalf on the Internet, such as sending an email, getting information or shopping online? Exclude work related tasks.
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} once or twice
\item \code{3} more often
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8115/vshow}{GSS Data Explorer page for \code{intbhlf}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intcomm}},
\code{\link{intliteracy}},
\code{\link{intpub}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplgrsn}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
