% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{intcntct}
\alias{intcntct}
\title{How much of r's communication is via text, mobile phone, or internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
intcntct
}
\details{
Question Think now of your contact with all of your family members and close friends. How much of it is through text messages, mobile phones,
or other communication devices that use the internet?
}
\section{Values}{

\itemize{
\item \code{1} all or almost all of it
\item \code{2} most of it
\item \code{3} about half of it
\item \code{4} some of it
\item \code{5} none or almost none of it
\item \code{6} i do not use any of these devices
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7270/vshow}{GSS Data Explorer page for \code{intcntct}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab all or almost all of it \tab most of it \tab about half of it \tab some of it \tab none or almost none of it \tab i do not use any of these devices \tab can't choose \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 275 \tab 357 \tab 265 \tab 184 \tab 78 \tab 10 \tab 3 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 275 \tab 357 \tab 265 \tab 184 \tab 78 \tab 10 \tab 3 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
