% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{internet}
\alias{internet}
\title{Internet makes info available more people}
\source{
General Social Survey https://gss.norc.org
}
\description{
internet
}
\details{
Question How much do you agree or disagree with the following statements? (Please, select one response)
G. A benefit of the Internet is that it makes information available to more and more people worldwide.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4889/vshow}{GSS Data Explorer page for \code{internet}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 574 \tab 519 \tab 79 \tab 20 \tab 6 \tab 17 \tab 1 \tab 2812 \cr
   Total \tab 1596 \tab 574 \tab 519 \tab 79 \tab 20 \tab 6 \tab 17 \tab 1 \tab 2812 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta2a}},
\code{\link{nafta3}}
}
\concept{ISSP National Identity}
\keyword{variable}
