% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{inthisp}
\alias{inthisp}
\title{Is interviewer spanish, hispanic or latino}
\source{
General Social Survey https://gss.norc.org
}
\description{
inthisp
}
\details{
Question Interviewer, are you Spanish, Hispanic, or Latino/Latina?  Which group are you from: Mexican, Mexican American, Chicano/Chicana, Puerto Rican, Cuban, Other. Please specific other.
}
\section{Values}{

\itemize{
\item \code{1} not hispanic
\item \code{2} mexican, mexican american, chicano/a
\item \code{3} puerto rican
\item \code{4} cuban
\item \code{5} salvadorian
\item \code{6} guatemalan
\item \code{7} panamanian
\item \code{8} nicaraguan
\item \code{9} costa rican
\item \code{10} central american
\item \code{11} honduran
\item \code{15} dominican
\item \code{16} west indian
\item \code{20} peruvian
\item \code{21} ecuadorian
\item \code{22} columbian
\item \code{23} venezuelan
\item \code{24} argentinian
\item \code{27} colombian
\item \code{30} spanish
\item \code{31} basque
\item \code{35} filipino/a
\item \code{40} latin american
\item \code{41} south american
\item \code{45} latin
\item \code{46} latino/a
\item \code{47} hispanic
\item \code{50} other, not specified
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5511/vshow}{GSS Data Explorer page for \code{inthisp}}.
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp1a}},
\code{\link{bthgrp1b}},
\code{\link{bthgrp1c}},
\code{\link{bthgrp1d}},
\code{\link{bthgrp1e}},
\code{\link{bthgrp1f}},
\code{\link{bthgrp2a}},
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp2f}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}},
\code{\link{bthgrp3f}},
\code{\link{bthgrp4a}},
\code{\link{bthgrp4b}},
\code{\link{bthgrp4c}},
\code{\link{bthgrp4d}},
\code{\link{bthgrp4e}},
\code{\link{bthgrp4f}},
\code{\link{bthgrp5a}},
\code{\link{bthgrp5b}},
\code{\link{bthgrp5c}},
\code{\link{bthgrp5d}},
\code{\link{bthgrp5e}},
\code{\link{bthgrp5f}},
\code{\link{frstgrp1}},
\code{\link{frstgrp2}},
\code{\link{frstgrp3}},
\code{\link{frstgrp4}},
\code{\link{frstgrp5}},
\code{\link{intrace1}},
\code{\link{intrace2}},
\code{\link{intrace3}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
