% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_673.R
\name{inthome}
\alias{inthome}
\title{Frequency r has home internet access}
\source{
General Social Survey https://gss.norc.org
}
\description{
inthome
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{6} no access
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
