% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{inthrss}
\alias{inthrss}
\title{Has r or someone they know been harassed online}
\source{
General Social Survey https://gss.norc.org
}
\description{
inthrss
}
\details{
Question Some people report experiences of harassment on the Internet. Have you or someone you know personally ever experienced online harassment or online hate speech?
}
\section{Values}{

\itemize{
\item \code{1} yes, me myself
\item \code{2} yes, someone i know personally
\item \code{3} yes, me and someone else i know personally
\item \code{4} no
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8125/vshow}{GSS Data Explorer page for \code{inthrss}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intscam}},
\code{\link{intviews}},
\code{\link{nextgen1}},
\code{\link{polnewsfrom}},
\code{\link{smnews}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
