% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{intintl}
\alias{intintl}
\title{Interested in international issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
intintl
}
\details{
Question International and foreign policy issues. Are you very interested, moderately interested, or not at all interested?
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} moderately interested
\item \code{3} not at all interested
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3452/vshow}{GSS Data Explorer page for \code{intintl}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{scimode}}

Other United States: 
\code{\link{intfarm}}
}
\concept{Foreign Affairs}
\concept{Science Knowledge & Attitudes}
\concept{United States}
\keyword{variable}
