% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_675.R
\name{intlnly}
\alias{intlnly}
\title{R would feel lonely without internet?}
\source{
General Social Survey https://gss.norc.org
}
\description{
intlnly
}
\details{
Question How much do you agree or disagree with the following statements?… Without the Internet I would feel lonely.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8054/vshow}{GSS Data Explorer page for \code{intlnly}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intcnfrm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intrust}},
\code{\link{intsearch}},
\code{\link{intstream}},
\code{\link{inttrig}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
