% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_382.R
\name{intmil}
\alias{intmil}
\title{Interested in military policy}
\source{
General Social Survey https://gss.norc.org
}
\description{
intmil
}
\details{
Question Issues about military and defense policy.  (Are you very interested, moderately interested, or not at all interested?)
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} moderately interested
\item \code{3} not at all interested
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3461/vshow}{GSS Data Explorer page for \code{intmil}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very interested \tab moderately interested \tab not at all interested \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 611 \tab 674 \tab 210 \tab 9 \tab 1 \tab 2023 \cr
   2010 \tab 1334 \tab 259 \tab 341 \tab 101 \tab 3 \tab 6 \tab 2044 \cr
   2012 \tab 972 \tab 397 \tab 448 \tab 144 \tab 7 \tab 6 \tab 1974 \cr
   2014 \tab 1299 \tab 439 \tab 614 \tab 183 \tab 2 \tab 1 \tab 2538 \cr
   2016 \tab 1477 \tab 489 \tab 662 \tab 234 \tab 2 \tab 3 \tab 2867 \cr
   2018 \tab 1173 \tab 385 \tab 577 \tab 211 \tab 2 \tab - \tab 2348 \cr
   Total \tab 6773 \tab 2580 \tab 3316 \tab 1083 \tab 25 \tab 17 \tab 13794 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intenvir}},
\code{\link{intspace}},
\code{\link{sciintro}},
\code{\link{scimath}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
