% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_675.R
\name{intnews}
\alias{intnews}
\title{How often r accesses non-routine political news sources}
\source{
General Social Survey https://gss.norc.org
}
\description{
intnews
}
\details{
Question When looking for news or information about political matters online, how often, if at all, do you… Check a news source that’s different from what you normally read, watch or listen to online?
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} often
\item \code{3} sometimes
\item \code{4} rarely
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8183/vshow}{GSS Data Explorer page for \code{intnews}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intcnfrm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{intlnly}},
\code{\link{intmeet}},
\code{\link{intrust}},
\code{\link{intsearch}},
\code{\link{intstream}},
\code{\link{inttrig}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
