% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{intoppo}
\alias{intoppo}
\title{Trend of online contact with politically different}
\source{
General Social Survey https://gss.norc.org
}
\description{
intoppo
}
\details{
Question In the last 12 months, if at all, has your online contact with the following people increased, decreased or remained the same?… People with different political views from your own.
}
\section{Values}{

\itemize{
\item \code{1} greatly decreased
\item \code{2} somewhat decreased
\item \code{3} remained the same
\item \code{4} somewhat increased
\item \code{5} greatly increased
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8122/vshow}{GSS Data Explorer page for \code{intoppo}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{inthrss}},
\code{\link{intpolview}},
\code{\link{intscam}},
\code{\link{intviews}},
\code{\link{nextgen1}},
\code{\link{polnewsfrom}},
\code{\link{smnews}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
