% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{intpol}
\alias{intpol}
\title{How interested in politics is r}
\source{
General Social Survey https://gss.norc.org
}
\description{
intpol
}
\details{
Question How interested are you in politics and national affairs?  Are you very interested, somewhat interested, only slightly interested, or not at all interested?
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} somewhat int
\item \code{3} slightly int
\item \code{4} not interested
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1027/vshow}{GSS Data Explorer page for \code{intpol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very interested \tab somewhat int \tab slightly int \tab not interested \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 402 \tab 742 \tab 481 \tab 180 \tab 5 \tab 9 \tab 1819 \cr
   Total \tab 0 \tab 402 \tab 742 \tab 481 \tab 180 \tab 5 \tab 9 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
