% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_675.R
\name{intpolview}
\alias{intpolview}
\title{Trend of online contact with politically similar}
\source{
General Social Survey https://gss.norc.org
}
\description{
intpolview
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} greatly decreased
\item \code{2} somewhat decreased
\item \code{3} remained the same
\item \code{4} somewhat increased
\item \code{5} greatly increased
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
