% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_602.R
\name{intrace2}
\alias{intrace2}
\title{Interviewer s race2}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrace2
}
\details{
Question What is your race?  Indicate one or more races that you consider yourself to be.
B. Second mention.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black or african-american
\item \code{3} american indian or alaska native
\item \code{4} asian indian
\item \code{5} chinese
\item \code{6} filipino
\item \code{7} japanese
\item \code{8} korean
\item \code{9} vietnamese
\item \code{10} other asian
\item \code{11} native hawaiian
\item \code{12} guamanian or chamorro
\item \code{13} samoan
\item \code{14} other pacific islander
\item \code{15} some other race
\item \code{16} hispanic
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5513/vshow}{GSS Data Explorer page for \code{intrace2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab white \tab black or african-american \tab american indian or alaska native \tab korean \tab other asian \tab some other race \tab hispanic \tab no answer \tab chinese \tab native hawaiian \tab don't know \tab filipino \tab other pacific islander \tab japanese \tab vietnamese \tab asian indian \tab not available in this release \tab Total \cr
   2004 \tab 2749 \tab 3 \tab 6 \tab 27 \tab 1 \tab 1 \tab 2 \tab 5 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4276 \tab 68 \tab 5 \tab 99 \tab - \tab - \tab 2 \tab 41 \tab 18 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1943 \tab 38 \tab 2 \tab 38 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1891 \tab 33 \tab 13 \tab 93 \tab - \tab - \tab 6 \tab - \tab - \tab 1 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1841 \tab 69 \tab 5 \tab 52 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2450 \tab 14 \tab 14 \tab 12 \tab - \tab - \tab 1 \tab 43 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2522 \tab 163 \tab 13 \tab 42 \tab - \tab - \tab 17 \tab 49 \tab 14 \tab - \tab - \tab 3 \tab 1 \tab - \tab 41 \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab 2146 \tab 101 \tab 10 \tab 7 \tab - \tab - \tab 26 \tab - \tab 43 \tab 2 \tab - \tab 12 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 19818 \tab 489 \tab 68 \tab 370 \tab 1 \tab 1 \tab 60 \tab 138 \tab 93 \tab 5 \tab 7 \tab 19 \tab 2 \tab 1 \tab 41 \tab 2 \tab 1 \tab 4032 \tab 25148 \cr
}
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp1a}},
\code{\link{bthgrp1b}},
\code{\link{bthgrp1c}},
\code{\link{bthgrp1d}},
\code{\link{bthgrp1e}},
\code{\link{bthgrp1f}},
\code{\link{bthgrp2a}},
\code{\link{intrace1}},
\code{\link{intrace3}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
