% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_602.R
\name{intrace2}
\alias{intrace2}
\title{Interviewer s race2}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrace2
}
\details{
Question What is your race?  Indicate one or more races that you consider yourself to be.
B. Second mention.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black or african-american
\item \code{3} american indian or alaska native
\item \code{4} asian indian
\item \code{5} chinese
\item \code{6} filipino
\item \code{7} japanese
\item \code{8} korean
\item \code{9} vietnamese
\item \code{10} other asian
\item \code{11} native hawaiian
\item \code{12} guamanian or chamorro
\item \code{13} samoan
\item \code{14} other pacific islander
\item \code{15} some other race
\item \code{16} hispanic
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5513/vshow}{GSS Data Explorer page for \code{intrace2}}.
}

\seealso{
Other Interviewer Characteristics: 
\code{\link{bthgrp1a}},
\code{\link{bthgrp1b}},
\code{\link{bthgrp1c}},
\code{\link{bthgrp1d}},
\code{\link{bthgrp1e}},
\code{\link{bthgrp1f}},
\code{\link{bthgrp2a}},
\code{\link{bthgrp2b}},
\code{\link{bthgrp2c}},
\code{\link{bthgrp2d}},
\code{\link{bthgrp2e}},
\code{\link{bthgrp2f}},
\code{\link{bthgrp3a}},
\code{\link{bthgrp3b}},
\code{\link{bthgrp3c}},
\code{\link{bthgrp3d}},
\code{\link{bthgrp3e}},
\code{\link{bthgrp3f}},
\code{\link{bthgrp4a}},
\code{\link{bthgrp4b}},
\code{\link{bthgrp4c}},
\code{\link{bthgrp4d}},
\code{\link{bthgrp4e}},
\code{\link{bthgrp4f}},
\code{\link{bthgrp5a}},
\code{\link{bthgrp5b}},
\code{\link{bthgrp5c}},
\code{\link{bthgrp5d}},
\code{\link{bthgrp5e}},
\code{\link{bthgrp5f}},
\code{\link{frstgrp1}},
\code{\link{frstgrp2}},
\code{\link{frstgrp3}},
\code{\link{frstgrp4}},
\code{\link{frstgrp5}},
\code{\link{inthisp}},
\code{\link{intrace1}},
\code{\link{intrace3}}
}
\concept{Interviewer Characteristics}
\keyword{variable}
