% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{intrnetuse}
\alias{intrnetuse}
\title{Frequency of internet use over past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
intrnetuse
}
\details{
Question During the last 12 months, how often, if at all, did you typically use the Internet on any device whether for work or personal use? If you are uncertain about your usage, please select the answer that comes closest to your use.
}
\section{Values}{

\itemize{
\item \code{1} almost all the time
\item \code{2} several times a day
\item \code{3} once a day
\item \code{4} several times a week
\item \code{5} several times a month
\item \code{6} less often
\item \code{7} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8124/vshow}{GSS Data Explorer page for \code{intrnetuse}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{nouseint1}},
\code{\link{nouseint2}},
\code{\link{nouseint3}},
\code{\link{nouseint4}},
\code{\link{nouseint5}},
\code{\link{nouseint6}},
\code{\link{nouseint7}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
