% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{intscam}
\alias{intscam}
\title{Has r or someone they know been a victim of online scam}
\source{
General Social Survey https://gss.norc.org
}
\description{
intscam
}
\details{
Question Some people receive emails or text messages from scammers from time to time, some of them also become victims. Apart from only receiving such messages, have you or someone you know personally ever been a victim of an online scam, identity theft or hacking attack?
}
\section{Values}{

\itemize{
\item \code{1} yes, me myself
\item \code{2} yes, someone i know personally
\item \code{3} yes, me and someone else i know personally
\item \code{4} no
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8126/vshow}{GSS Data Explorer page for \code{intscam}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{inthrss}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intviews}},
\code{\link{nextgen1}},
\code{\link{polnewsfrom}},
\code{\link{smnews}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
