% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_285.R
\name{intuse}
\alias{intuse}
\title{Use internet/apps more than occassionally}
\source{
General Social Survey https://gss.norc.org
}
\description{
intuse
}
\details{
Question Do you use the internet or web-enabled applications, known as APPS, for more than email, at least occasionally?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6266/vshow}{GSS Data Explorer page for \code{intuse}}.
}

\seealso{
Other Social Media Usage: 
\code{\link{intmbile}}
}
\concept{Social Media Usage}
\keyword{variable}
