% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{intwkdyh}
\alias{intwkdyh}
\title{Hours of internet use on weekdays}
\source{
General Social Survey https://gss.norc.org
}
\description{
intwkdyh
}
\details{
Question HOURS:
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6288/vshow}{GSS Data Explorer page for \code{intwkdyh}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 2 \tab 20 \tab 24 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1630 \tab 150 \tab 341 \tab 30 \tab 2 \tab 22 \tab 3 \tab 4 \tab 9 \tab 1 \tab 250 \tab 4 \tab 8 \tab 130 \tab 83 \tab 64 \tab 52 \tab 14 \tab 41 \tab 12 \tab 7 \tab 10 \tab 2867 \cr
   Total \tab 1630 \tab 150 \tab 341 \tab 30 \tab 2 \tab 22 \tab 3 \tab 4 \tab 9 \tab 1 \tab 250 \tab 4 \tab 8 \tab 130 \tab 83 \tab 64 \tab 52 \tab 14 \tab 41 \tab 12 \tab 7 \tab 10 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{intwkenm}},
\code{\link{snsmfrst}}
}
\concept{Social Media Usage}
\keyword{variable}
