% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{intwkdym}
\alias{intwkdym}
\title{Minutes of internet use on weekdays}
\source{
General Social Survey https://gss.norc.org
}
\description{
intwkdym
}
\details{
Question MINUTES:
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6287/vshow}{GSS Data Explorer page for \code{intwkdym}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 15 \tab 2 \tab 20 \tab 25 \tab 3 \tab 30 \tab 40 \tab 45 \tab 5 \tab 55 \tab 59 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1630 \tab 1024 \tab 1 \tab 12 \tab 14 \tab 1 \tab 15 \tab 1 \tab 1 \tab 137 \tab 3 \tab 11 \tab 2 \tab 1 \tab 1 \tab 5 \tab 8 \tab 2867 \cr
   Total \tab 1630 \tab 1024 \tab 1 \tab 12 \tab 14 \tab 1 \tab 15 \tab 1 \tab 1 \tab 137 \tab 3 \tab 11 \tab 2 \tab 1 \tab 1 \tab 5 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{intwkdyh}},
\code{\link{intwkenh}},
\code{\link{intwkenm}},
\code{\link{snsmfrst}}
}
\concept{Social Media Usage}
\keyword{variable}
