% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{intwkenh}
\alias{intwkenh}
\title{Hours of internet use on weekends}
\source{
General Social Survey https://gss.norc.org
}
\description{
intwkenh
}
\details{
Question HOURS:
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6290/vshow}{GSS Data Explorer page for \code{intwkenh}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 13 \tab 14 \tab 15 \tab 18 \tab 2 \tab 20 \tab 24 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1630 \tab 189 \tab 329 \tab 24 \tab 22 \tab 3 \tab 6 \tab 6 \tab 1 \tab 241 \tab 7 \tab 7 \tab 136 \tab 115 \tab 61 \tab 39 \tab 8 \tab 20 \tab 2 \tab 10 \tab 11 \tab 2867 \cr
   Total \tab 1630 \tab 189 \tab 329 \tab 24 \tab 22 \tab 3 \tab 6 \tab 6 \tab 1 \tab 241 \tab 7 \tab 7 \tab 136 \tab 115 \tab 61 \tab 39 \tab 8 \tab 20 \tab 2 \tab 10 \tab 11 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenm}},
\code{\link{snsmfrst}}
}
\concept{Social Media Usage}
\keyword{variable}
