% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{intwkenm}
\alias{intwkenm}
\title{Minutes of internet use on weekends}
\source{
General Social Survey https://gss.norc.org
}
\description{
intwkenm
}
\details{
Question MINUTES:
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6289/vshow}{GSS Data Explorer page for \code{intwkenm}}.
}

\seealso{
Other Social Media Usage: 
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{snsmfrst}}
}
\concept{Social Media Usage}
\keyword{variable}
