% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{intyrs}
\alias{intyrs}
\title{Years of service as an interviewer at norc}
\source{
General Social Survey https://gss.norc.org
}
\description{
intyrs
}
\details{
Question D. Years of service as an interviewer at NORC
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018 \cr
   A/B/C/D \tab 2006 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{lngthinv}},
\code{\link{mode}()}
}
\concept{Paradata}
\keyword{variable}
