% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{intyrs}
\alias{intyrs}
\title{Years of service as an interviewer at norc}
\source{
General Social Survey https://gss.norc.org
}
\description{
intyrs
}
\details{
Question D. Years of service as an interviewer at NORC
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{lngthinv}},
\code{\link{mode}()}

Other Interview: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{lngthinv}},
\code{\link{mode}()}
}
\concept{Interview}
\concept{Paradata}
\keyword{variable}
