% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{israel}
\alias{israel}
\title{Liking for israel}
\source{
General Social Survey https://gss.norc.org
}
\description{
israel
}
\details{
Question You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? READ EACH ITEM.
G. Israel
}
\section{Values}{

\itemize{
\item \verb{[0]} +5
\item \verb{[1]} +4
\item \verb{[2]} +3
\item \verb{[3]} +2
\item \verb{[4]} +1
\item \verb{[5]} -1
\item \verb{[6]} -2
\item \verb{[7]} -3
\item \verb{[8]} -4
\item \verb{[9]} -5
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1974, 1975, 1977, 1982, 1983, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5496/vshow}{GSS Data Explorer page for \code{israel}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab +5 \tab +4 \tab +3 \tab +2 \tab +1 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab iap \tab Total \cr
   1974 \tab - \tab 198 \tab 141 \tab 231 \tab 175 \tab 258 \tab 128 \tab 57 \tab 67 \tab 33 \tab 89 \tab 107 \tab - \tab 1484 \cr
   1975 \tab - \tab 146 \tab 123 \tab 205 \tab 169 \tab 275 \tab 164 \tab 82 \tab 77 \tab 48 \tab 76 \tab 125 \tab - \tab 1490 \cr
   1977 \tab - \tab 180 \tab 151 \tab 204 \tab 173 \tab 270 \tab 149 \tab 71 \tab 64 \tab 41 \tab 69 \tab 158 \tab - \tab 1530 \cr
   1982 \tab - \tab 207 \tab 147 \tab 250 \tab 211 \tab 290 \tab 215 \tab 121 \tab 92 \tab 62 \tab 95 \tab 170 \tab - \tab 1860 \cr
   1983 \tab - \tab 148 \tab 123 \tab 169 \tab 172 \tab 285 \tab 209 \tab 106 \tab 111 \tab 67 \tab 103 \tab 106 \tab - \tab 1599 \cr
   1985 \tab - \tab 168 \tab 125 \tab 211 \tab 201 \tab 254 \tab 160 \tab 94 \tab 82 \tab 48 \tab 75 \tab 116 \tab - \tab 1534 \cr
   1986 \tab - \tab 165 \tab 128 \tab 175 \tab 180 \tab 285 \tab 144 \tab 84 \tab 73 \tab 53 \tab 104 \tab 79 \tab - \tab 1470 \cr
   1988 \tab - \tab 96 \tab 60 \tab 101 \tab 91 \tab 155 \tab 130 \tab 77 \tab 77 \tab 30 \tab 94 \tab 570 \tab - \tab 1481 \cr
   1989 \tab - \tab 87 \tab 56 \tab 107 \tab 95 \tab 177 \tab 104 \tab 70 \tab 59 \tab 43 \tab 118 \tab 621 \tab - \tab 1537 \cr
   1990 \tab - \tab 75 \tab 48 \tab 86 \tab 93 \tab 182 \tab 128 \tab 69 \tab 45 \tab 30 \tab 93 \tab 523 \tab - \tab 1372 \cr
   1991 \tab - \tab 152 \tab 104 \tab 162 \tab 138 \tab 177 \tab 76 \tab 38 \tab 35 \tab 14 \tab 55 \tab 566 \tab - \tab 1517 \cr
   1993 \tab - \tab 105 \tab 73 \tab 133 \tab 104 \tab 227 \tab 123 \tab 71 \tab 59 \tab 41 \tab 66 \tab 604 \tab - \tab 1606 \cr
   1994 \tab - \tab 46 \tab 32 \tab 52 \tab 59 \tab 110 \tab 55 \tab 31 \tab 34 \tab 14 \tab 34 \tab 2525 \tab - \tab 2992 \cr
   Total \tab 0 \tab 1773 \tab 1311 \tab 2086 \tab 1861 \tab 2945 \tab 1785 \tab 971 \tab 875 \tab 524 \tab 1071 \tab 6270 \tab 0 \tab 21472 \cr
}
}

\seealso{
Other War and International: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{japan}}
}
\concept{War and International}
\keyword{variable}
