% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{jbendmo}
\alias{jbendmo}
\title{Month last paid job ended}
\source{
General Social Survey https://gss.norc.org
}
\description{
jbendmo
}
\details{
Question Please also specify the month.        In (MONTHS):
}
\section{Values}{

\itemize{
\item \code{1} january
\item \code{2} february
\item \code{3} march
\item \code{4} april
\item \code{5} may
\item \code{6} june
\item \code{7} july
\item \code{8} august
\item \code{9} september
\item \code{10} october
\item \code{11} november
\item \code{12} december
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6340/vshow}{GSS Data Explorer page for \code{jbendmo}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab january \tab february \tab march \tab april \tab may \tab june \tab july \tab august \tab september \tab october \tab november \tab december \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2382 \tab 31 \tab 23 \tab 24 \tab 31 \tab 53 \tab 62 \tab 36 \tab 23 \tab 23 \tab 26 \tab 22 \tab 30 \tab 75 \tab 26 \tab 2867 \cr
   Total \tab 2382 \tab 31 \tab 23 \tab 24 \tab 31 \tab 53 \tab 62 \tab 36 \tab 23 \tab 23 \tab 26 \tab 22 \tab 30 \tab 75 \tab 26 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
