% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{jbendyr}
\alias{jbendyr}
\title{Year last paid job ended}
\source{
General Social Survey https://gss.norc.org
}
\description{
jbendyr
}
\details{
Question When did your last paid job end?        In (YEAR):
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6339/vshow}{GSS Data Explorer page for \code{jbendyr}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1936 \tab 1944 \tab 1965 \tab 1967 \tab 1969 \tab 1970 \tab 1973 \tab 1976 \tab 1978 \tab 1980 \tab 1981 \tab 1984 \tab 1986 \tab 1987 \tab 1988 \tab 1989 \tab 1990 \tab 1991 \tab 1992 \tab 1993 \tab 1994 \tab 1995 \tab 1996 \tab 1997 \tab 1998 \tab 1999 \tab 2000 \tab 2001 \tab 2002 \tab 2003 \tab 2004 \tab 2005 \tab 2006 \tab 2007 \tab 2008 \tab 2009 \tab 2010 \tab 2011 \tab 2012 \tab 2013 \tab 2014 \tab 2015 \tab 2016 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2382 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 6 \tab 10 \tab 6 \tab 13 \tab 9 \tab 7 \tab 12 \tab 16 \tab 8 \tab 12 \tab 11 \tab 20 \tab 10 \tab 18 \tab 22 \tab 19 \tab 25 \tab 13 \tab 29 \tab 23 \tab 43 \tab 41 \tab 40 \tab 15 \tab 25 \tab 2867 \cr
   Total \tab 2382 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 6 \tab 10 \tab 6 \tab 13 \tab 9 \tab 7 \tab 12 \tab 16 \tab 8 \tab 12 \tab 11 \tab 20 \tab 10 \tab 18 \tab 22 \tab 19 \tab 25 \tab 13 \tab 29 \tab 23 \tab 43 \tab 41 \tab 40 \tab 15 \tab 25 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
