% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{jbintfam}
\alias{jbintfam}
\title{How often r felt: demand of job interfere with family life}
\source{
General Social Survey https://gss.norc.org
}
\description{
jbintfam
}
\details{
Question How often do you feel that the demands of your job interfere with your family life?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4614/vshow}{GSS Data Explorer page for \code{jbintfam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 51 \tab 101 \tab 364 \tab 293 \tab 206 \tab 1 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 38 \tab 109 \tab 308 \tab 324 \tab 156 \tab - \tab 5 \tab 2867 \cr
   Total \tab 5419 \tab 89 \tab 210 \tab 672 \tab 617 \tab 362 \tab 1 \tab 7 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
