% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{jbtrain}
\alias{jbtrain}
\title{Did r have any training to improve job skill past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
jbtrain
}
\details{
Question Over the past 12 months, have you had any training to improve your job skills, either at the workplace or somewhere else?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4617/vshow}{GSS Data Explorer page for \code{jbtrain}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 610 \tab 405 \tab 1 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 555 \tab 379 \tab - \tab 6 \tab 2867 \cr
   Total \tab 5419 \tab 1165 \tab 784 \tab 1 \tab 8 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
