% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{jew}
\alias{jew}
\title{What sort of jew r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
jew
}
\details{
Question B. Do you consider yourself Orthodox, Conservative, Reform or none of these? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \code{1} orthodox
\item \code{2} conservative
\item \code{3} reform
\item \code{4} none of these
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/291/vshow}{GSS Data Explorer page for \code{jew}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab orthodox \tab conservative \tab reform \tab none of these \tab no answer \tab don't know \tab other \tab skipped on web \tab Total \cr
   1988 \tab 1450 \tab 1 \tab 16 \tab 10 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1509 \tab - \tab 7 \tab 11 \tab 5 \tab 5 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1340 \tab 1 \tab 12 \tab 9 \tab 4 \tab 5 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1483 \tab 1 \tab 13 \tab 11 \tab 6 \tab 3 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1573 \tab - \tab 10 \tab 13 \tab 10 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2922 \tab 3 \tab 14 \tab 30 \tab 12 \tab 11 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2836 \tab 4 \tab 17 \tab 29 \tab 15 \tab 3 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2752 \tab 4 \tab 12 \tab 21 \tab 11 \tab 32 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2751 \tab 3 \tab 11 \tab 28 \tab 11 \tab 13 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2702 \tab 4 \tab 7 \tab 23 \tab 13 \tab 15 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 2745 \tab 5 \tab 11 \tab 28 \tab 10 \tab 13 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4406 \tab 10 \tab 19 \tab 33 \tab 14 \tab 28 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1976 \tab 7 \tab 14 \tab 12 \tab 6 \tab 8 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1995 \tab 2 \tab 8 \tab 18 \tab 9 \tab 12 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1944 \tab 1 \tab 3 \tab 16 \tab 6 \tab 4 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2483 \tab 4 \tab 7 \tab 15 \tab 11 \tab 16 \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2801 \tab 5 \tab 9 \tab 20 \tab 16 \tab 15 \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2293 \tab 3 \tab 8 \tab 16 \tab 9 \tab 16 \tab 1 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 3881 \tab 2 \tab 9 \tab 43 \tab 21 \tab 44 \tab - \tab - \tab 32 \tab 4032 \cr
   2022 \tab 3442 \tab 7 \tab 9 \tab 22 \tab 16 \tab 33 \tab - \tab 6 \tab 9 \tab 3544 \cr
   Total \tab 49284 \tab 67 \tab 216 \tab 408 \tab 208 \tab 277 \tab 3 \tab 11 \tab 41 \tab 50515 \cr
}
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jewaj}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{spattend}}
}
\concept{Core}
\keyword{variable}
