% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{jew16aj}
\alias{jew16aj}
\title{What sort of jew r was at age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
jew16aj
}
\details{
Question Were you raised as a Jew?
A. Were you Orthodox, Conservative, Reform, or none of these? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \verb{[1]} orthodox
\item \verb{[2]} conservative
\item \verb{[3]} reform
\item \verb{[4]} none of these
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/346/vshow}{GSS Data Explorer page for \code{jew16aj}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab reform \tab no answer \tab Total \cr
   2012 \tab 1969 \tab 1 \tab 4 \tab 1974 \cr
   Total \tab 1969 \tab 1 \tab 4 \tab 1974 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom16}},
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{oth16}},
\code{\link{othjew16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}
}
\concept{Core}
\keyword{variable}
