% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{jew_next}
\alias{jew_next}
\title{What sort of jew - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
jew_next
}
\details{
Question  Do you consider yourself Orthodox, Conservative, Reform or none of these?
}
\section{Values}{

\itemize{
\item \code{1} orthodox
\item \code{2} conservative
\item \code{3} reform
\item \code{4} none of these
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8003/vshow}{GSS Data Explorer page for \code{jew_next}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab orthodox \tab conservative \tab reform \tab none of these \tab other \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1190 \tab 5 \tab 7 \tab 15 \tab 9 \tab 5 \tab 2308 \tab 5 \tab 3544 \cr
   Total \tab 1190 \tab 5 \tab 7 \tab 15 \tab 9 \tab 5 \tab 2308 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
