% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{jewaj}
\alias{jewaj}
\title{What sort of jew r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
jewaj
}
\details{
Question B. Do you consider yourself Orthodox, Conservative, Reform or none of these? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \code{1} orthodox
\item \code{2} conservative
\item \code{3} reform
\item \code{4} none of these
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/292/vshow}{GSS Data Explorer page for \code{jewaj}}.
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jew}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{relitena}},
\code{\link{spattend}}
}
\concept{Core}
\keyword{variable}
