% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_256.R
\name{jewschng}
\alias{jewschng}
\title{Percentage change population of jews in 25 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
jewschng
}
\details{
Question In the next 25 years, which of these groups do you think will increase their share of the population by a lot, which will increase by a little, which will stay the same, which will decrease by a little, and which will decrease by a lot?
C. Jews?
}
\section{Values}{

\itemize{
\item \verb{[1]} increase a lot
\item \verb{[2]} increase by a little
\item \verb{[3]} stay the same
\item \verb{[4]} decrease by a little
\item \verb{[5]} decrease by a lot
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2319/vshow}{GSS Data Explorer page for \code{jewschng}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increase a lot \tab increase by a little \tab stay the same \tab decrease by a little \tab decrease by a lot \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 50 \tab 328 \tab 695 \tab 163 \tab 25 \tab 124 \tab 13 \tab 2817 \cr
   Total \tab 1419 \tab 50 \tab 328 \tab 695 \tab 163 \tab 25 \tab 124 \tab 13 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkchng}},
\code{\link{hispchng}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{whtchng}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
