% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_259.R
\name{jewswrk}
\alias{jewswrk}
\title{Does r know jews from work}
\source{
General Social Survey https://gss.norc.org
}
\description{
jewswrk
}
\details{
Question Now, I’m going to ask some questions about these people you personally know.
D. Do you know any of these Jews from the place where you work?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2348/vshow}{GSS Data Explorer page for \code{jewswrk}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkwrk}},
\code{\link{hispschl}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}}

Other Jews: 
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}}
}
\concept{Jews}
\concept{Multi-Ethnic United States}
\keyword{variable}
