% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{jobaff}
\alias{jobaff}
\title{Favor preferences in hiring blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobaff
}
\details{
Question Some people say that because of past discrimination, blacks should be given preference in hiring and promotion. Others say that such preference in hiring and promotion of blacks is wrong because it discriminates against others. What about your opinion--are you for or against preferential hiring and promotion of blacks?
IF FOR:Do you favor preference in hiring and promotion strongly or not strongly?
IF AGAINST:Do you oppose preference in hiring and promotion strongly or not strongly?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly for pref
\item \verb{[2]} for preference
\item \verb{[3]} against preference
\item \verb{[4]} strongly against pref
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1625/vshow}{GSS Data Explorer page for \code{jobaff}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly for pref \tab for preference \tab against preference \tab strongly against pref \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 130 \tab 91 \tab 276 \tab 887 \tab 78 \tab 12 \tab 2992 \cr
   Total \tab 1518 \tab 130 \tab 91 \tab 276 \tab 887 \tab 78 \tab 12 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{ethhist}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
