% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{jobcntct}
\alias{jobcntct}
\title{Times r sought job information from contact outside}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobcntct
}
\details{
Question Now I'd like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit: Click on a link in an email message from a friend, relative, or co-worker.
E. Business or work contacts outside your workplace–that is outside where you were then working?
}
\section{Values}{

\itemize{
\item \code{0} not at all
\item \code{1} 1 or 2 times
\item \code{2} 3-5 times
\item \code{3} 6 or more times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2614/vshow}{GSS Data Explorer page for \code{jobcntct}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab 6 or more times \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2220 \tab 216 \tab 186 \tab 85 \tab 51 \tab 1 \tab 6 \tab 2765 \cr
   2004 \tab 2453 \tab 163 \tab 101 \tab 48 \tab 39 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 4673 \tab 379 \tab 287 \tab 133 \tab 90 \tab 2 \tab 13 \tab 5577 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emailyr}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webable}},
\code{\link{webyr}}
}
\concept{Information Society}
\keyword{variable}
