% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{jobcntrl}
\alias{jobcntrl}
\title{Who makes decision for your job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobcntrl
}
\details{
Question I'm going to read four statements. (HAND CARD BV TO RESPONDENT) Please tell me which one best describes your job:
}
\section{Values}{

\itemize{
\item \verb{[1]} other decide what and how i do
\item \verb{[2]} other decide what but not how
\item \verb{[3]} i decide some what and how i do
\item \verb{[4]} i am my own boss
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1307/vshow}{GSS Data Explorer page for \code{jobcntrl}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab other decide what and how i do \tab other decide what but not how \tab i decide some what and how i do \tab i am my own boss \tab no answer \tab Total \cr
   1991 \tab 605 \tab 86 \tab 214 \tab 328 \tab 267 \tab 17 \tab 1517 \cr
   Total \tab 605 \tab 86 \tab 214 \tab 328 \tab 267 \tab 17 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}
}
\concept{Work Organizations}
\keyword{variable}
