% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{jobfind1}
\alias{jobfind1}
\title{How easy for r to find a same job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobfind1
}
\details{
Question How easy would it be for you to find a job with another employer with approximately the same income and fringe benefits as you have now?
}
\section{Values}{

\itemize{
\item \code{1} very easy to find similar job
\item \code{2} somewhat easy to find similar job
\item \code{3} not easy at all to find similar job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2819/vshow}{GSS Data Explorer page for \code{jobfind1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very easy to find similar job \tab somewhat easy to find similar job \tab not easy at all to find similar job \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 427 \tab 626 \tab 702 \tab 25 \tab 16 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 503 \tab 636 \tab 552 \tab 31 \tab 12 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 170 \tab 398 \tab 573 \tab 22 \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 224 \tab 475 \tab 533 \tab 13 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 341 \tab 592 \tab 469 \tab 14 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 547 \tab 833 \tab 558 \tab 17 \tab 16 \tab 18 \tab 3544 \cr
   Total \tab 8345 \tab 2212 \tab 3560 \tab 3387 \tab 122 \tab 105 \tab 18 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}},
\code{\link{wksexism}}
}
\concept{Quality of Working Life}
\keyword{variable}
