% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{jobinc}
\alias{jobinc}
\title{High income}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobinc
}
\details{
Question A. Would you please look at this card and tell me which one thing on this list you would most prefer in a job?
B. Which comes next?
C. Which is third most important?
D. Which is fourth most important?
A. High Income
}
\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   2006 \tab D/- \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/550/vshow}{GSS Data Explorer page for \code{jobinc}}.
}

\seealso{
Other Core: 
\code{\link{anomia7}},
\code{\link{anomia8}},
\code{\link{anomia9}},
\code{\link{jobfind}},
\code{\link{joblose}},
\code{\link{jobsec}},
\code{\link{richwork}},
\code{\link{satjob}}
}
\concept{Core}
\keyword{variable}
