% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_142.R
\name{jobjudge}
\alias{jobjudge}
\title{Someone evaluate your performance regularly}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobjudge
}
\details{
Question Does someone formally evaluate your job performance on a regular basis?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1335/vshow}{GSS Data Explorer page for \code{jobjudge}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 577 \tab 319 \tab 3 \tab 13 \tab 1517 \cr
   Total \tab 605 \tab 577 \tab 319 \tab 3 \tab 13 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{totmoney}},
\code{\link{wrkwell}},
\code{\link{youmoney}}
}
\concept{Work Organizations}
\keyword{variable}
