% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{joborfam}
\alias{joborfam}
\title{Would you remain in unsatisfying job for family}
\source{
General Social Survey https://gss.norc.org
}
\description{
joborfam
}
\details{
Question (HAND CARD D3)
Have you ever remained or would you remain in a job that was not satisfying for you for the benefit of your family life?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, i have done so and probably would do so again
\item \verb{[2]} yes, i have done so but probably would not do so again
\item \verb{[3]} no, i have not done so but probably would do so
\item \verb{[4]} no, i have not done so and probably would not do so
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6324/vshow}{GSS Data Explorer page for \code{joborfam}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, i have done so and probably would do so again \tab yes, i have done so but probably would not do so again \tab no, i have not done so but probably would do so \tab no, i have not done so and probably would not do so \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 564 \tab 232 \tab 431 \tab 235 \tab 5 \tab 10 \tab 2867 \cr
   Total \tab 1390 \tab 564 \tab 232 \tab 431 \tab 235 \tab 5 \tab 10 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
