% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_62.R
\name{jobpromo}
\alias{jobpromo}
\title{Chances for advancement}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobpromo
}
\details{
Question A. Would you please look at this card and tell me which one thing on this list you would most prefer in a job?
B. Which comes next?
C. Which is third most important?
D. Which is fourth most important?
D. Chances for advancement
}
\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/B \tab 2012, 2014 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/553/vshow}{GSS Data Explorer page for \code{jobpromo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab most impt \tab second \tab third \tab fourth \tab fifth \tab no answer \tab Total \cr
   1973 \tab - \tab 266 \tab 517 \tab 342 \tab 195 \tab 142 \tab 42 \tab 1504 \cr
   1974 \tab - \tab 261 \tab 549 \tab 332 \tab 193 \tab 119 \tab 30 \tab 1484 \cr
   1976 \tab - \tab 257 \tab 529 \tab 308 \tab 217 \tab 138 \tab 50 \tab 1499 \cr
   1977 \tab - \tab 296 \tab 507 \tab 342 \tab 204 \tab 131 \tab 50 \tab 1530 \cr
   1980 \tab - \tab 272 \tab 502 \tab 363 \tab 194 \tab 105 \tab 32 \tab 1468 \cr
   1982 \tab 937 \tab 156 \tab 256 \tab 193 \tab 182 \tab 95 \tab 41 \tab 1860 \cr
   1984 \tab - \tab 277 \tab 516 \tab 343 \tab 197 \tab 108 \tab 32 \tab 1473 \cr
   1985 \tab - \tab 334 \tab 546 \tab 344 \tab 189 \tab 87 \tab 34 \tab 1534 \cr
   1987 \tab - \tab 334 \tab 615 \tab 424 \tab 265 \tab 134 \tab 47 \tab 1819 \cr
   1988 \tab 504 \tab 191 \tab 322 \tab 221 \tab 132 \tab 80 \tab 31 \tab 1481 \cr
   1989 \tab 504 \tab 170 \tab 361 \tab 244 \tab 139 \tab 66 \tab 53 \tab 1537 \cr
   1990 \tab 455 \tab 156 \tab 312 \tab 219 \tab 127 \tab 66 \tab 37 \tab 1372 \cr
   1991 \tab 524 \tab 154 \tab 329 \tab 231 \tab 152 \tab 79 \tab 48 \tab 1517 \cr
   1993 \tab 531 \tab 177 \tab 358 \tab 257 \tab 163 \tab 82 \tab 38 \tab 1606 \cr
   1994 \tab 2508 \tab 82 \tab 149 \tab 114 \tab 77 \tab 41 \tab 21 \tab 2992 \cr
   2006 \tab 2992 \tab 283 \tab 439 \tab 352 \tab 265 \tab 159 \tab 20 \tab 4510 \cr
   2012 \tab 1002 \tab 115 \tab 246 \tab 233 \tab 194 \tab 158 \tab 26 \tab 1974 \cr
   2014 \tab 1288 \tab 163 \tab 325 \tab 300 \tab 248 \tab 196 \tab 18 \tab 2538 \cr
   Total \tab 11245 \tab 3944 \tab 7378 \tab 5162 \tab 3333 \tab 1986 \tab 650 \tab 33698 \cr
}
}

\seealso{
Other Core: 
\code{\link{jobhonor}},
\code{\link{jobhour}},
\code{\link{jobindep}},
\code{\link{jobinter}},
\code{\link{jobkeep}},
\code{\link{jobmeans}},
\code{\link{joboff}},
\code{\link{jobpay}},
\code{\link{jobrise}}
}
\concept{Core}
\keyword{variable}
