% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{jobsall}
\alias{jobsall}
\title{Govts resp: provide jobs for all}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobsall
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to . . .
A. Provide a job for everyone who wants one.
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4076/vshow}{GSS Data Explorer page for \code{jobsall}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
