% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{jobsat}
\alias{jobsat}
\title{How satisfied is r with his/her job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobsat
}
\details{
Question How satisfied are you in your job?
}
\section{Values}{

\itemize{
\item \code{1} completely satisfied
\item \code{2} very satisfied
\item \code{3} fairly satisfied
\item \code{4} neither satisfied nor dissatisfied
\item \code{5} fairly dissatisfied
\item \code{6} very dissatisfied
\item \code{7} completely dissatisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4561/vshow}{GSS Data Explorer page for \code{jobsat}}.
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}

Other Job: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Job}
\concept{Satisfaction}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
