% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{jobsecok}
\alias{jobsecok}
\title{The job security is good}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobsecok
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
K. The job security is good
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2012, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2806/vshow}{GSS Data Explorer page for \code{jobsecok}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 951 \tab 546 \tab 162 \tab 108 \tab 15 \tab 14 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 909 \tab 540 \tab 151 \tab 90 \tab 26 \tab 18 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 560 \tab 400 \tab 98 \tab 91 \tab 12 \tab 26 \tab - \tab 2044 \cr
   2012 \tab 795 \tab 615 \tab 410 \tab 68 \tab 47 \tab 10 \tab 29 \tab - \tab 1974 \cr
   2014 \tab 1288 \tab 679 \tab 411 \tab 90 \tab 58 \tab 4 \tab 8 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 859 \tab 412 \tab 89 \tab 45 \tab 8 \tab 35 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 1100 \tab 667 \tab 98 \tab 80 \tab 7 \tab 15 \tab 22 \tab 3544 \cr
   Total \tab 9140 \tab 5673 \tab 3386 \tab 756 \tab 519 \tab 82 \tab 145 \tab 22 \tab 19723 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{cowrkint}},
\code{\link{fringeok}},
\code{\link{haveinfo}},
\code{\link{hlpequip}},
\code{\link{opdevel}},
\code{\link{promtefr}},
\code{\link{supcares}},
\code{\link{wkfreedm}}

Other Workplace Conflict: 
\code{\link{supcares}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
