% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{jobskill}
\alias{jobskill}
\title{How important training in developing skills}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobskill
}
\details{
Question Now think about the skills that you actually use in your job. How important would you say each of the following was in developing these skills?
B. Training or experience gained in my present or a previous job
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} not impt/unimp
\item \verb{[4]} not important
\item \verb{[5]} not imprt at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4584/vshow}{GSS Data Explorer page for \code{jobskill}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab not impt/unimp \tab not important \tab not imprt at all \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 408 \tab 294 \tab 71 \tab 23 \tab 26 \tab 7 \tab 36 \tab 2832 \cr
   Total \tab 1967 \tab 408 \tab 294 \tab 71 \tab 23 \tab 26 \tab 7 \tab 36 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
