% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{jobtrain}
\alias{jobtrain}
\title{Obligation to train for civilian jobs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobtrain
}
\details{
Question Most people in the Armed Forces are taught skills they can use in civilian jobs later. But some don't get such training. They are taught only combat skills. Do you think the Armed Forces have an obligation to train everybody in service For civilian jobs later, or is that not a responsibility of the Armed Forces?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, obligation
\item \verb{[2]} not obligation
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/821/vshow}{GSS Data Explorer page for \code{jobtrain}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, obligation \tab not obligation \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 690 \tab 722 \tab 36 \tab 25 \tab 1473 \cr
   Total \tab 0 \tab 690 \tab 722 \tab 36 \tab 25 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{defwrkev}},
\code{\link{guerilla}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()},
\code{\link{vetaid}}
}
\concept{Core}
\keyword{variable}
