% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_487.R
\name{jobvsfam}
\alias{jobvsfam}
\title{Difficult to fulfill fam responsibility bc of job}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobvsfam
}
\details{
Question How often has each of the following happened to you during the past three months?
B. It has been difficult for me to fulfill my family responsibilities because of the amount of time I spent on my job.
}
\section{Values}{

\itemize{
\item \code{1} several times a week
\item \code{2} several times a month
\item \code{3} once or twice
\item \code{4} never
\item \code{6} doesnt apply or no job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4448/vshow}{GSS Data Explorer page for \code{jobvsfam}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{buythngs}},
\code{\link{famvswrk}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{weekend}}

Other Work: 
\code{\link{famvswrk}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}}
}
\concept{Family}
\concept{ISSP Family & Gender Roles}
\concept{Work}
\keyword{variable}
