% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{judgeluv}
\alias{judgeluv}
\title{Contrast between judge and lover}
\source{
General Social Survey https://gss.norc.org
}
\description{
judgeluv
}
\details{
Question There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? (HAND CARD AE) The first set of contrasting images shows Mother at 1 on the scale and Father at 7. If you imagine God as a Mother you would place yourself at 1. If you imagine God as a Father, you would place yourself at 7. If you imagine God as somewhere between Mother and Father, you would place yourself at 2, 3, 4, 5, or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F, SUBSTITUTING IMAGES A-F FOR "MOTHER" AND "FATHER".) Where would you place your image of God on the scale for . . . READ EACH SET OF IMAGES AND CIRCLE ONE CODE FOR EACH.
}
\section{Values}{

\itemize{
\item \code{1} judge
\item \code{7} lover
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/336/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab judge \tab lover \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 102 \tab 84 \tab 218 \tab 41 \tab 22 \tab 1 \tab 373 \tab 83 \tab 65 \tab - \tab 1473 \cr
   1985 \tab - \tab 118 \tab 113 \tab 363 \tab 81 \tab 51 \tab 20 \tab 583 \tab 148 \tab 57 \tab - \tab 1534 \cr
   1986 \tab - \tab 133 \tab 97 \tab 316 \tab 53 \tab 53 \tab 13 \tab 620 \tab 129 \tab 56 \tab - \tab 1470 \cr
   1987 \tab - \tab 154 \tab 122 \tab 449 \tab 60 \tab 53 \tab 20 \tab 706 \tab 169 \tab 86 \tab - \tab 1819 \cr
   1988 \tab - \tab 143 \tab 99 \tab 401 \tab 48 \tab 54 \tab 33 \tab 528 \tab 113 \tab 62 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 90 \tab 86 \tab 279 \tab 31 \tab 30 \tab 6 \tab 367 \tab 73 \tab 44 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 81 \tab 72 \tab 229 \tab 34 \tab 24 \tab 9 \tab 345 \tab 84 \tab 50 \tab - \tab 1372 \cr
   1991 \tab 158 \tab 140 \tab 116 \tab 320 \tab 75 \tab 55 \tab - \tab 385 \tab 59 \tab 209 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 103 \tab 76 \tab 250 \tab 54 \tab 46 \tab 9 \tab 396 \tab 87 \tab 59 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 197 \tab 137 \tab 485 \tab 77 \tab 61 \tab 31 \tab 737 \tab 172 \tab 80 \tab - \tab 2992 \cr
   1996 \tab 1913 \tab 85 \tab 77 \tab 268 \tab 32 \tab 31 \tab 12 \tab 351 \tab 95 \tab 40 \tab - \tab 2904 \cr
   1998 \tab 1387 \tab 135 \tab 108 \tab 420 \tab 56 \tab 59 \tab 22 \tab 482 \tab 106 \tab 57 \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 94 \tab 128 \tab 355 \tab 57 \tab 53 \tab 111 \tab 448 \tab 82 \tab 37 \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 51402 \tab 1575 \tab 1315 \tab 4353 \tab 699 \tab 592 \tab 287 \tab 6321 \tab 1400 \tab 902 \tab 6853 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
}
}

\seealso{
Other Core: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}},
\code{\link{rdeemlib}}

Other Religion: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}},
\code{\link{rdeemlib}}

Other Split Ballots: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{rdeemlib}}
}
\concept{Core}
\concept{Religion}
\concept{Split Ballots}
\keyword{variable}
