% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{justpay}
\alias{justpay}
\title{Is r's pay just}
\source{
General Social Survey https://gss.norc.org
}
\description{
justpay
}
\details{
Question Given your skills and effort, is your pay:
}
\section{Values}{

\itemize{
\item \verb{[1]} much less than is just
\item \verb{[2]} a little less than is just
\item \verb{[3]} about just for me
\item \verb{[4]} a little more than is just
\item \verb{[5]} much more than is just
\item \verb{[6]} never had a job
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4335/vshow}{GSS Data Explorer page for \code{justpay}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab much less than is just \tab a little less than is just \tab about just for me \tab a little more than is just \tab much more than is just \tab never had a job \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 237 \tab 432 \tab 386 \tab 58 \tab 14 \tab 15 \tab 60 \tab 70 \tab 2817 \cr
   Total \tab 1545 \tab 237 \tab 432 \tab 386 \tab 58 \tab 14 \tab 15 \tab 60 \tab 70 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
