% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{justpay}
\alias{justpay}
\title{Is r's pay just}
\source{
General Social Survey https://gss.norc.org
}
\description{
justpay
}
\details{
Question Given your skills and effort, is your pay:
}
\section{Values}{

\itemize{
\item \code{1} much less than is just
\item \code{2} a little less than is just
\item \code{3} about just for me
\item \code{4} a little more than is just
\item \code{5} much more than is just
\item \code{6} never had a job
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4335/vshow}{GSS Data Explorer page for \code{justpay}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
