% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_420.R
\name{kd1jwoth}
\alias{kd1jwoth}
\title{R's child 1 considered jewish}
\source{
General Social Survey https://gss.norc.org
}
\description{
kd1jwoth
}
\details{
Question Is (CHILD’S NAME) considered Jewish for any reason?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} half/part
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3838/vshow}{GSS Data Explorer page for \code{kd1jwoth}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab half/part \tab no \tab Total \cr
   2008 \tab 1866 \tab 2 \tab 3 \tab 152 \tab 2023 \cr
   2012 \tab 1967 \tab 2 \tab 1 \tab 4 \tab 1974 \cr
   Total \tab 3833 \tab 4 \tab 4 \tab 156 \tab 3997 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{kd2jwoth}},
\code{\link{kd3jwoth}},
\code{\link{kd4jwoth}},
\code{\link{kd5jwoth}}
}
\concept{Jewish Identity}
\keyword{variable}
