% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_420.R
\name{kd4jwoth}
\alias{kd4jwoth}
\title{R's child 4 considered jewish}
\source{
General Social Survey https://gss.norc.org
}
\description{
kd4jwoth
}
\details{
Question Is (CHILD'S NAME) considered Jewish for any reason?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} half/part
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3841/vshow}{GSS Data Explorer page for \code{kd4jwoth}}.

Counts by year:\tabular{llll}{
   year \tab iap \tab no \tab Total \cr
   2008 \tab 2014 \tab 9 \tab 2023 \cr
   Total \tab 2014 \tab 9 \tab 2023 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{kd1jwoth}},
\code{\link{kd2jwoth}},
\code{\link{kd3jwoth}},
\code{\link{kd5jwoth}}
}
\concept{Jewish Identity}
\keyword{variable}
