% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_419.R
\name{kd4relig}
\alias{kd4relig}
\title{Religion of r's child 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
kd4relig
}
\details{
Question In what religion is (CHILD’S NAME) being raised? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} christian
\item \code{12} native american
\item \code{13} inter-nondenominational
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3833/vshow}{GSS Data Explorer page for \code{kd4relig}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab buddhism \tab muslim/islam \tab christian \tab no answer \tab Total \cr
   2008 \tab 1960 \tab 33 \tab 13 \tab 3 \tab 9 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2023 \cr
   Total \tab 1960 \tab 33 \tab 13 \tab 3 \tab 9 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{bmitzvah}},
\code{\link{kd1relig}},
\code{\link{kd2relig}},
\code{\link{kd3relig}},
\code{\link{kd5relig}},
\code{\link{kd6relig}},
\code{\link{kd7relig}},
\code{\link{kd8relig}},
\code{\link{synmem}}
}
\concept{Jewish Identity}
\keyword{variable}
