% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_164.R
\name{kdevwork}
\alias{kdevwork}
\title{Did r child work for 1 or more yr?}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdevwork
}
\details{
Question Has (s/he) ever worked at one job for as long as one year?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1540/vshow}{GSS Data Explorer page for \code{kdevwork}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2845 \tab 81 \tab 52 \tab 7 \tab 7 \tab 2992 \cr
   Total \tab 2845 \tab 81 \tab 52 \tab 7 \tab 7 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
