% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_165.R
\name{kdindstatus}
\alias{kdindstatus}
\title{Kdind10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdindstatus
}
\details{
Question KDIND10 Coding Status
}
\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{8} case not collected in 1994
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6015/vshow}{GSS Data Explorer page for \code{kdindstatus}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab found, iap \tab found, no answer \tab found, coded \tab not found, iap \tab not found, coded \tab not found, no code assigned \tab Total \cr
   1994 \tab - \tab 2338 \tab 39 \tab 561 \tab 48 \tab 4 \tab 2 \tab 2992 \cr
   Total \tab 0 \tab 2338 \tab 39 \tab 561 \tab 48 \tab 4 \tab 2 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdindtag}},
\code{\link{kdpres10}},
\code{\link{kdpres105plus}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}}
}
\concept{Family Mobility}
\keyword{variable}
