% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_164.R
\name{kdoccindv}
\alias{kdoccindv}
\title{Kid's occ/ind verbatim retrieved}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdoccindv
}
\details{
Question Child's OCC/IND verbatim recovered
}
\section{Values}{

\itemize{
\item \code{0} no verbatims for kd
\item \code{1} verbatims for kd
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6011/vshow}{GSS Data Explorer page for \code{kdoccindv}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no verbatims for kd \tab verbatims for kd \tab Total \cr
   1994 \tab 54 \tab 2343 \tab 595 \tab 2992 \cr
   Total \tab 54 \tab 2343 \tab 595 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdpres80}},
\code{\link{kdwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
