% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_163.R
\name{kdpicked}
\alias{kdpicked}
\title{Number of child randomly chosen}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdpicked
}
\details{
Question NUMBER OF CHILD SELECTED FOR FOLLOW-UP QUESTIONS:
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1537/vshow}{GSS Data Explorer page for \code{kdpicked}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab Total \cr
   1994 \tab 2334 \tab 322 \tab 189 \tab 73 \tab 38 \tab 18 \tab 12 \tab 2 \tab 1 \tab 3 \tab 2992 \cr
   Total \tab 2334 \tab 322 \tab 189 \tab 73 \tab 38 \tab 18 \tab 12 \tab 2 \tab 1 \tab 3 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdeduc3}},
\code{\link{kdeduc4}},
\code{\link{kdeduc5}},
\code{\link{kdeduc6}},
\code{\link{kdeduc7}},
\code{\link{kdeduc8}},
\code{\link{kdeduc9}},
\code{\link{kdwork1}},
\code{\link{kdwork2}}
}
\concept{Family Mobility}
\keyword{variable}
