% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_165.R
\name{kdpres10}
\alias{kdpres10}
\title{R child occupational prestige score (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdpres10
}
\details{
Question Prestige of Respondent's Child's Occupation (2010)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6017/vshow}{GSS Data Explorer page for \code{kdpres10}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 16 \tab 17 \tab 18 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 80 \tab Total \cr
   1994 \tab 2428 \tab 2 \tab 2 \tab 1 \tab 9 \tab 7 \tab 1 \tab 7 \tab 22 \tab 4 \tab 2 \tab 15 \tab 3 \tab 2 \tab 40 \tab 14 \tab 16 \tab 4 \tab 32 \tab 12 \tab 9 \tab 28 \tab 14 \tab 9 \tab 5 \tab 6 \tab 11 \tab 29 \tab 27 \tab 37 \tab 28 \tab 16 \tab 11 \tab 9 \tab 5 \tab 12 \tab 4 \tab 3 \tab 2 \tab 3 \tab 3 \tab 14 \tab 18 \tab 5 \tab 13 \tab 18 \tab 5 \tab 3 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 3 \tab 5 \tab 1 \tab 1 \tab 2992 \cr
   Total \tab 2428 \tab 2 \tab 2 \tab 1 \tab 9 \tab 7 \tab 1 \tab 7 \tab 22 \tab 4 \tab 2 \tab 15 \tab 3 \tab 2 \tab 40 \tab 14 \tab 16 \tab 4 \tab 32 \tab 12 \tab 9 \tab 28 \tab 14 \tab 9 \tab 5 \tab 6 \tab 11 \tab 29 \tab 27 \tab 37 \tab 28 \tab 16 \tab 11 \tab 9 \tab 5 \tab 12 \tab 4 \tab 3 \tab 2 \tab 3 \tab 3 \tab 14 \tab 18 \tab 5 \tab 13 \tab 18 \tab 5 \tab 3 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 3 \tab 5 \tab 1 \tab 1 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdpres105plus}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}}
}
\concept{Family Mobility}
\keyword{variable}
