% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_165.R
\name{kdpres10}
\alias{kdpres10}
\title{R child occupational prestige score (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdpres10
}
\details{
Question Prestige of Respondent's Child's Occupation (2010)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6017/vshow}{GSS Data Explorer page for \code{kdpres10}}.
}

\seealso{
Other Family Mobility: 
\code{\link{kdindstatus}},
\code{\link{kdindtag}},
\code{\link{kdpres105plus}},
\code{\link{sbsex1}},
\code{\link{sbsex2}},
\code{\link{sbsex3}},
\code{\link{sbsex4}},
\code{\link{sbsex5}},
\code{\link{sbsex6}}
}
\concept{Family Mobility}
\keyword{variable}
