% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_164.R
\name{kdpres80}
\alias{kdpres80}
\title{R child occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdpres80
}
\details{
Question Respondent's Child's Occupation, Prestige, and Industry
A. What kind of work (is/was) (CHILD'S NAME) doing? That is, what (is/was) (his/her) job called?
B. IF NOT ALREADY ANSWERED, ASK:  What (does/did) (he/she) actually do in that job?  Tell me what (are/were) some of (his/her) main duties?
C. What kind of place (does/did) (s/he) work for?
D. IF NOT ALREADY ANSWERED, ASK:  What (do/did) they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  (Is/was) (s/he) self-employed or (does/did) (s/he) work for someone else?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1542/vshow}{GSS Data Explorer page for \code{kdpres80}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 17 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 70 \tab 71 \tab 73 \tab 74 \tab 75 \tab 86 \tab dk, na, iap \tab iap \tab Total \cr
   1994 \tab - \tab 5 \tab 3 \tab 4 \tab 1 \tab 11 \tab 17 \tab 17 \tab 2 \tab 7 \tab 23 \tab 23 \tab 13 \tab 19 \tab 17 \tab 11 \tab 11 \tab 17 \tab 21 \tab 4 \tab 2 \tab 12 \tab 15 \tab 8 \tab 20 \tab 8 \tab 15 \tab 8 \tab 25 \tab 24 \tab 5 \tab 20 \tab 8 \tab 35 \tab 13 \tab 3 \tab 4 \tab 4 \tab 6 \tab 2 \tab 2 \tab 7 \tab 14 \tab 1 \tab 2 \tab 14 \tab 12 \tab 19 \tab 4 \tab 5 \tab 1 \tab 1 \tab 2 \tab 10 \tab 4 \tab 1 \tab 2430 \tab - \tab 2992 \cr
   Total \tab 0 \tab 5 \tab 3 \tab 4 \tab 1 \tab 11 \tab 17 \tab 17 \tab 2 \tab 7 \tab 23 \tab 23 \tab 13 \tab 19 \tab 17 \tab 11 \tab 11 \tab 17 \tab 21 \tab 4 \tab 2 \tab 12 \tab 15 \tab 8 \tab 20 \tab 8 \tab 15 \tab 8 \tab 25 \tab 24 \tab 5 \tab 20 \tab 8 \tab 35 \tab 13 \tab 3 \tab 4 \tab 4 \tab 6 \tab 2 \tab 2 \tab 7 \tab 14 \tab 1 \tab 2 \tab 14 \tab 12 \tab 19 \tab 4 \tab 5 \tab 1 \tab 1 \tab 2 \tab 10 \tab 4 \tab 1 \tab 2430 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
