% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_164.R
\name{kdpres80}
\alias{kdpres80}
\title{R child occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdpres80
}
\details{
Question Respondent's Child's Occupation, Prestige, and Industry
A. What kind of work (is/was) (CHILD'S NAME) doing? That is, what (is/was) (his/her) job called?
B. IF NOT ALREADY ANSWERED, ASK:  What (does/did) (he/she) actually do in that job?  Tell me what (are/were) some of (his/her) main duties?
C. What kind of place (does/did) (s/he) work for?
D. IF NOT ALREADY ANSWERED, ASK:  What (do/did) they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  (Is/was) (s/he) self-employed or (does/did) (s/he) work for someone else?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1542/vshow}{GSS Data Explorer page for \code{kdpres80}}.
}

\seealso{
Other Family Mobility: 
\code{\link{kdevwork}},
\code{\link{kdind10}},
\code{\link{kdind80}},
\code{\link{kdocc10}},
\code{\link{kdocc80}},
\code{\link{kdoccindv}},
\code{\link{kdoccstatus}},
\code{\link{kdocctag}},
\code{\link{kdwrkslf}}

Other Children: 
\code{\link{kdevwork}},
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdwrkslf}}

Other Occupation: 
\code{\link{kdind80}},
\code{\link{kdocc80}},
\code{\link{kdwrkslf}}
}
\concept{Children}
\concept{Family Mobility}
\concept{Occupation}
\concept{Occupational Prestige}
\keyword{variable}
