% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_161.R
\name{kdrel8}
\alias{kdrel8}
\title{R relation to r 8th child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdrel8
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
D8. Relation to R. of eighth child
}
\section{Values}{

\itemize{
\item \code{1} biological
\item \code{2} adopted
\item \code{3} step
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1517/vshow}{GSS Data Explorer page for \code{kdrel8}}.
}

\seealso{
Other Family Mobility: 
\code{\link{kdalive1}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel9}}

Other Children: 
\code{\link{kdalive1}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel9}}

Other Family: 
\code{\link{kdalive1}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel9}}

Other Household: 
\code{\link{kdalive1}},
\code{\link{kdrel1}},
\code{\link{kdrel2}},
\code{\link{kdrel3}},
\code{\link{kdrel4}},
\code{\link{kdrel5}},
\code{\link{kdrel6}},
\code{\link{kdrel7}},
\code{\link{kdrel9}}
}
\concept{Children}
\concept{Family}
\concept{Family Mobility}
\concept{Household}
\keyword{variable}
