% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_292.R
\name{kdtmkid3}
\alias{kdtmkid3}
\title{Time child 3 spends with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdtmkid3
}
\details{
Question Would you say that (NAME OF CHILD) feels that the amount of time they spend with you is too much, about the right amount, or too little?
C. Third listed child?
}
\section{Values}{

\itemize{
\item \verb{[1]} too much
\item \verb{[2]} about the right amount
\item \verb{[3]} too little
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2649/vshow}{GSS Data Explorer page for \code{kdtmkid3}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too much \tab about the right amount \tab too little \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2748 \tab 7 \tab 32 \tab 17 \tab 1 \tab 12 \tab 2817 \cr
   Total \tab 2748 \tab 7 \tab 32 \tab 17 \tab 1 \tab 12 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{knowkid2}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}}
}
\concept{Information Society}
\keyword{variable}
