% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_160.R
\name{kdyrbrn9}
\alias{kdyrbrn9}
\title{Birth year of r 9th child}
\source{
General Social Survey https://gss.norc.org
}
\description{
kdyrbrn9
}
\details{
Question Now I have some questions about your children. These include your biological, adopted, or step children. Let's start with the oldest and work down to the youngest. (RECORD THIS INFORMATION FOR UP TO THE NINE OLDEST CHILDREN IN TABLE TO RIGHT.)
C9. Birth year of ninth child.
}
\section{Values}{

\itemize{
\item \code{9996} before 1976
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1509/vshow}{GSS Data Explorer page for \code{kdyrbrn9}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1933 \tab 1949 \tab 1950 \tab 1961 \tab 1963 \tab 1966 \tab 1968 \tab 1973 \tab Total \cr
   1994 \tab 2983 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2992 \cr
   Total \tab 2983 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex9}},
\code{\link{kdyrbrn1}},
\code{\link{kdyrbrn2}},
\code{\link{kdyrbrn3}},
\code{\link{kdyrbrn4}},
\code{\link{kdyrbrn5}},
\code{\link{kdyrbrn6}},
\code{\link{kdyrbrn7}},
\code{\link{kdyrbrn8}}
}
\concept{Family Mobility}
\keyword{variable}
