% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{kid5up}
\alias{kid5up}
\title{Have children older than 5}
\source{
General Social Survey https://gss.norc.org
}
\description{
kid5up
}
\details{
Question Do you have any children older than 5?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1146/vshow}{GSS Data Explorer page for \code{kid5up}}.
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{givearts}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}},
\code{\link{relhrs2}}
}
\concept{Religion}
\keyword{variable}
