% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_481.R
\name{kidjoy}
\alias{kidjoy}
\title{Kids are life's greatest joy}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidjoy
}
\details{
Question Do you agree or disagree ...
B. Watching children grow up is life''s greatest joy.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4381/vshow}{GSS Data Explorer page for \code{kidjoy}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidless}},
\code{\link{kidnofre}}

Other Children: 
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidless}},
\code{\link{kidnofre}}
}
\concept{Children}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
