% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{kidmost}
\alias{kidmost}
\title{Do r have the most contact with son or daughter}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidmost
}
\details{
Question Of your children aged 18 and older, with whom do you have the most contact?
}
\section{Values}{

\itemize{
\item \verb{[1]} with a son
\item \verb{[2]} with a daughter
\item \verb{[3]} i have no contact with any adult children
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4189/vshow}{GSS Data Explorer page for \code{kidmost}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab with a son \tab with a daughter \tab i have no contact with any adult children \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2307 \tab 179 \tab 259 \tab 9 \tab 3 \tab 8 \tab 2765 \cr
   Total \tab 2307 \tab 179 \tab 259 \tab 9 \tab 3 \tab 8 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
