% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{kidnow1}
\alias{kidnow1}
\title{R's children under 18 living with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidnow1
}
\details{
Question Are your biological or adopted children under age 18 living with you now at least some of the time?
}
\section{Values}{

\itemize{
\item \code{1} yes, all of them live with me at least some of the time
\item \code{2} some live with me at least some of the time
\item \code{3} none live with me
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3917/vshow}{GSS Data Explorer page for \code{kidnow1}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidresp}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{toldsxor}}
}
\concept{Sexual Orientation}
\keyword{variable}
