% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{kidnow2}
\alias{kidnow2}
\title{R is living in a hh where children under 18 are living with r under r's care}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidnow2
}
\details{
Question Right now are you living in a household where any children under 18 are living with you at least some of the time and are under your care and responsibility?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3918/vshow}{GSS Data Explorer page for \code{kidnow2}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{losejob5}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Children}
\concept{Family}
\concept{Household}
\concept{Sexual Orientation}
\keyword{variable}
