% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{kidnum}
\alias{kidnum}
\title{Number of children age 18 and older}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidnum
}
\details{
Question Now some questions about your children who are aged 18 and older. How many children age 18 and older do you have? (We mean children who are still alive. Please include step-children and adopted children.)
If you have no children age 18 and older, enter '0'.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4188/vshow}{GSS Data Explorer page for \code{kidnum}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{Jewish Identity}
\keyword{variable}
