% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{kidnum}
\alias{kidnum}
\title{Number of children age 18 and older}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidnum
}
\details{
Question Now some questions about your children who are aged 18 and older. How many children age 18 and older do you have? (We mean children who are still alive. Please include step-children and adopted children.)
If you have no children age 18 and older, enter '0'.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4188/vshow}{GSS Data Explorer page for \code{kidnum}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 9 \tab Total \cr
   2002 \tab 1616 \tab 691 \tab 99 \tab 162 \tab 100 \tab 46 \tab 18 \tab 8 \tab 21 \tab 4 \tab 2765 \cr
   2012 \tab 1955 \tab - \tab 7 \tab 10 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   Total \tab 3571 \tab 691 \tab 106 \tab 172 \tab 102 \tab 46 \tab 18 \tab 8 \tab 21 \tab 4 \tab 4739 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{Jewish Identity}
\keyword{variable}
