% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{kidresp}
\alias{kidresp}
\title{R is legally responsible biological/adoptive parent of a child under 18}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidresp
}
\details{
Question Are you currently the legally responsible biological or adoptive parent of a child under age 18?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3916/vshow}{GSS Data Explorer page for \code{kidresp}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{toldsxor}}

Other Children: 
\code{\link{adoptkid}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}}

Other Family: 
\code{\link{adoptkid}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{marunion}}

Other Household: 
\code{\link{kidlived}},
\code{\link{kidnow1}}
}
\concept{Children}
\concept{Family}
\concept{Household}
\concept{Sexual Orientation}
\keyword{variable}
