% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{kidsinhh}
\alias{kidsinhh}
\title{Does interviewer believe that there are children under the age of 15 present at}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidsinhh
}
\details{
Question Based on the features of this housing unit and the surrounding property/neighborhood that you can see, do you believe that there are children under the age of 15 present in the household?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6655/vshow}{GSS Data Explorer page for \code{kidsinhh}}.
}

\keyword{variable}
