% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{kidsocst}
\alias{kidsocst}
\title{Having children increases social standing in society}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidsocst
}
\details{
Question To what extent do you agree or disagree with the following statements?
E. Having children increases people’s social standing in society.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4461/vshow}{GSS Data Explorer page for \code{kidsocst}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 46 \tab 367 \tab 381 \tab 391 \tab 63 \tab 49 \tab 5 \tab - \tab 1974 \cr
   2022 \tab 2380 \tab 36 \tab 259 \tab 446 \tab 276 \tab 63 \tab 70 \tab 5 \tab 9 \tab 3544 \cr
   Total \tab 3052 \tab 82 \tab 626 \tab 827 \tab 667 \tab 126 \tab 119 \tab 10 \tab 9 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{paidlv}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
