% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{kidssolv}
\alias{kidssolv}
\title{R's kids living standard compared to r (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidssolv
}
\details{
Question When your children are at the age you are now, do you think their standard of living will be much better, somewhat better, about the same, somewhat worse, or much worse than yours is now?
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} somewhat better
\item \code{3} about the same
\item \code{4} somewhat worse
\item \code{5} much worse
\item \code{6} no children (phone mode only; volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7599/vshow}{GSS Data Explorer page for \code{kidssolv}}.
}

\keyword{variable}
