% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{kidsuffr}
\alias{kidsuffr}
\title{Preschooler will suffer if mom works}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidsuffr
}
\details{
Question Do you agree or disagree ...
B. A pre-school child is likely to suffer if his or her mother works.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4344/vshow}{GSS Data Explorer page for \code{kidsuffr}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 142 \tab 452 \tab 175 \tab 400 \tab 208 \tab 15 \tab 22 \tab 1481 \cr
   1994 \tab 1545 \tab 130 \tab 458 \tab 175 \tab 459 \tab 195 \tab 14 \tab 16 \tab 2992 \cr
   2002 \tab 1594 \tab 137 \tab 305 \tab 285 \tab - \tab 423 \tab 20 \tab 1 \tab 2765 \cr
   2012 \tab 672 \tab 68 \tab 299 \tab 109 \tab 511 \tab 294 \tab 14 \tab 7 \tab 1974 \cr
   Total \tab 3878 \tab 477 \tab 1514 \tab 744 \tab 1370 \tab 1120 \tab 63 \tab 46 \tab 9212 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famsuffr}},
\code{\link{fejobind}},
\code{\link{hapifwrk}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{mawrkwrm}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
