% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{kidsund18}
\alias{kidsund18}
\title{Any kids under 18 in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidsund18
}
\details{
Question This section contains questions about you and your household. We are interested in understanding what households in the United States look like and we are grateful for any information you provide.
Are there any individuals aged 0 to 17 years old living at this household?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8015/vshow}{GSS Data Explorer page for \code{kidsund18}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1788 \tab 528 \tab 1213 \tab 10 \tab 5 \tab 3544 \cr
   Total \tab 1788 \tab 528 \tab 1213 \tab 10 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
