% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{kidtrble}
\alias{kidtrble}
\title{Kids more trouble than they are worth}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidtrble
}
\details{
Question Do you agree or disagree ...
A. Children are more trouble than they are worth.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4380/vshow}{GSS Data Explorer page for \code{kidtrble}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 18 \tab 39 \tab 127 \tab 550 \tab 643 \tab 20 \tab 17 \tab 1481 \cr
   Total \tab 67 \tab 18 \tab 39 \tab 127 \tab 550 \tab 643 \tab 20 \tab 17 \tab 1481 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{mardiv}},
\code{\link{marhomo}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{threkids}},
\code{\link{twokids}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
