% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_439.R
\name{kidxfilm}
\alias{kidxfilm}
\title{Govt intervene: chld allowed to see porn}
\source{
General Social Survey https://gss.norc.org
}
\description{
kidxfilm
}
\details{
Question Sometimes public authorities intervene with parents in raising their children. Please indicate in each of the following cases how far you think authorities should go in dealing with a 10 year old child and his or her parents. Public authorities should...
H. The parents allow the child to watch violent or pornographic films.
}
\section{Values}{

\itemize{
\item \code{1} take no action
\item \code{2} give warning
\item \code{3} take child
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4035/vshow}{GSS Data Explorer page for \code{kidxfilm}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{beltup}},
\code{\link{kidbeat}},
\code{\link{kiddrugs}},
\code{\link{kidedpar}},
\code{\link{kidhlth}},
\code{\link{kidneedy}},
\code{\link{kidout}},
\code{\link{kidskips}},
\code{\link{nosmoke}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
