% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{knomentl}
\alias{knomentl}
\title{Gov should have the right to know}
\source{
General Social Survey https://gss.norc.org
}
\description{
knomentl
}
\details{
Question Please tell me which one of the statements on the card best describes what the government has the right to know. Before giving an individual a SECRET or TOP SECRET clearance, the government should have the right to know:
}
\section{Values}{

\itemize{
\item \code{1} nothing about mental health
\item \code{2} whether currently consulting
\item \code{3} ever consulted a mental
\item \code{4} ever consulted & diagnosis
\item \code{5} diagnosis & confident. info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5226/vshow}{GSS Data Explorer page for \code{knomentl}}.
}

\seealso{
Other National Security: 
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{leakinfo}},
\code{\link{punarms}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}},
\code{\link{taketrck}}
}
\concept{National Security}
\keyword{variable}
