% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_292.R
\name{knowkid2}
\alias{knowkid2}
\title{R knows what child is doing on weekend}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowkid2
}
\details{
Question Thinking now only about weekends (SATURDAY AND SUNDAY) and only when your child is awake"would you say that you know exactly what (NAME OF CHILD) is doing on the weekend almost all of the time, most of the time, about half of the time, some of the time or hardly any of the time?
B. Second listed child?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost all of the time
\item \verb{[2]} most of the time
\item \verb{[3]} about half of the time
\item \verb{[4]} some of the time
\item \verb{[5]} hardly any of the time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2652/vshow}{GSS Data Explorer page for \code{knowkid2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost all of the time \tab most of the time \tab about half of the time \tab some of the time \tab hardly any of the time \tab no answer \tab Total \cr
   2000 \tab 2589 \tab 133 \tab 58 \tab 15 \tab 6 \tab 2 \tab 14 \tab 2817 \cr
   Total \tab 2589 \tab 133 \tab 58 \tab 15 \tab 6 \tab 2 \tab 14 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{kdtmkid1}},
\code{\link{kdtmkid2}},
\code{\link{kdtmkid3}},
\code{\link{kdtmkid4}},
\code{\link{knowkid1}},
\code{\link{sptmkid1}},
\code{\link{sptmkid2}},
\code{\link{sptmkid3}},
\code{\link{sptmkid4}}
}
\concept{Information Society}
\keyword{variable}
