% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_102.R
\name{known1}
\alias{known1}
\title{How long has r known person #1?}
\source{
General Social Survey https://gss.norc.org
}
\description{
known1
}
\details{
Question Have you known (NAME) for less than 3 years, 3 to 6 years, or more than 6 years?  ASK FOR EACH NAME.
\enumerate{
\item Known First Person
}
}
\section{Values}{

\itemize{
\item \code{1} lt 3 years
\item \code{2} 3-6 years
\item \code{3} 6 years +
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/935/vshow}{GSS Data Explorer page for \code{known1}}.
}

\seealso{
Other Social Networks: 
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}}

Other Social Networks: 
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}}
}
\concept{Social Networks}
\keyword{variable}
