% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_102.R
\name{known2}
\alias{known2}
\title{How long has r known person #2?}
\source{
General Social Survey https://gss.norc.org
}
\description{
known2
}
\details{
Question Have you known (NAME) for less than 3 years, 3 to 6 years, or more than 6 years?  ASK FOR EACH NAME.
2. Known Second Person
}
\section{Values}{

\itemize{
\item \code{1} lt 3 years
\item \code{2} 3-6 years
\item \code{3} 6 years +
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/936/vshow}{GSS Data Explorer page for \code{known2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab lt 3 years \tab 3-6 years \tab 6 years + \tab no answer \tab Total \cr
   1985 \tab 364 \tab 138 \tab 164 \tab 856 \tab 12 \tab 1534 \cr
   Total \tab 364 \tab 138 \tab 164 \tab 856 \tab 12 \tab 1534 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{known1}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}}
}
\concept{Social Networks}
\keyword{variable}
