% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{knowpols}
\alias{knowpols}
\title{Public official s treatment depends on who they know}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowpols
}
\details{
Question Do you think that the treatment people get from public officials in America depends on who they know?
}
\section{Values}{

\itemize{
\item \code{1} definitely does
\item \code{2} probably does
\item \code{3} probably does not
\item \code{4} definitely does not
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4122/vshow}{GSS Data Explorer page for \code{knowpols}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bribe}},
\code{\link{corrupt1}},
\code{\link{corrupt2}}

Other Political: 
\code{\link{bribe}},
\code{\link{corrupt1}},
\code{\link{corrupt2}}
}
\concept{ISSP Role of Gov't}
\concept{Political}
\keyword{variable}
