% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{knowschd}
\alias{knowschd}
\title{How far in advance know work schedule}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowschd
}
\details{
Question How far in advance do you usually know what days and hours you will need to work?
}
\section{Values}{

\itemize{
\item \code{1} one week or less
\item \code{2} between 1 and 2 weeks
\item \code{3} between 3 and 4 weeks
\item \code{4} 4 weeks or more
\item \code{5} my schedule never changes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2839/vshow}{GSS Data Explorer page for \code{knowschd}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab one week or less \tab between 1 and 2 weeks \tab between 3 and 4 weeks \tab 4 weeks or more \tab don't know \tab no answer \tab 5 \tab skipped on web \tab Total \cr
   2014 \tab 1288 \tab 540 \tab 165 \tab 65 \tab 459 \tab 13 \tab 8 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 599 \tab 183 \tab 85 \tab 531 \tab 20 \tab 30 \tab - \tab - \tab 2348 \cr
   2022 \tab 1540 \tab 440 \tab 233 \tab 118 \tab 165 \tab 18 \tab 12 \tab 1000 \tab 18 \tab 3544 \cr
   Total \tab 3728 \tab 1579 \tab 581 \tab 268 \tab 1155 \tab 51 \tab 50 \tab 1000 \tab 18 \tab 8430 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{satjob1}},
\code{\link{slpprblm}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}
}
\concept{Quality of Working Life}
\keyword{variable}
