% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{knowschd1}
\alias{knowschd1}
\title{How far in advance know work schedule (2022 version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowschd1
}
\details{
Question How far in advance do you usually know what days and hours you will need to work?
}
\section{Values}{

\itemize{
\item \verb{[1]} one week or less
\item \verb{[2]} between 1 and 2 weeks
\item \verb{[3]} between 3 and 4 weeks
\item \verb{[4]} 4 weeks or more
\item \verb{[5]} my schedule never changes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8041/vshow}{GSS Data Explorer page for \code{knowschd1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab not available in this year \tab one week or less \tab between 1 and 2 weeks \tab between 3 and 4 weeks \tab 4 weeks or more \tab my schedule never changes \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 440 \tab 233 \tab 118 \tab 165 \tab 1000 \tab 18 \tab 1540 \tab 12 \tab 18 \tab 3544 \cr
   Total \tab 0 \tab 440 \tab 233 \tab 118 \tab 165 \tab 1000 \tab 18 \tab 1540 \tab 12 \tab 18 \tab 3544 \cr
}
}

\concept{Quality of Working Life}
\keyword{variable}
