% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{knowschd1}
\alias{knowschd1}
\title{How far in advance know work schedule (2022 version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowschd1
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} one week or less
\item \code{2} between 1 and 2 weeks
\item \code{3} between 3 and 4 weeks
\item \code{4} 4 weeks or more
\item \code{5} my schedule never changes
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
