% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{knowschd1}
\alias{knowschd1}
\title{How far in advance know work schedule (2022 version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
knowschd1
}
\details{
Question How far in advance do you usually know what days and hours you will need to work?
}
\section{Values}{

\itemize{
\item \code{1} one week or less
\item \code{2} between 1 and 2 weeks
\item \code{3} between 3 and 4 weeks
\item \code{4} 4 weeks or more
\item \code{5} my schedule never changes
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8041/vshow}{GSS Data Explorer page for \code{knowschd1}}.
}

\concept{Quality of Working Life}
\keyword{variable}
