% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_258.R
\name{knwjew}
\alias{knwjew}
\title{Does r personally any jews}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwjew
}
\details{
Question Do you personally know any...?
C. Jews?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2332/vshow}{GSS Data Explorer page for \code{knwjew}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1455 \tab 752 \tab 559 \tab 1 \tab 50 \tab 2817 \cr
   Total \tab 1455 \tab 752 \tab 559 \tab 1 \tab 50 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkschl}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
