% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{knwmhosp}
\alias{knwmhosp}
\title{Know anyone who was in a hospital}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwmhosp
}
\details{
Question Did you ever know anyone who was in a hospital because of a mental illness?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2058/vshow}{GSS Data Explorer page for \code{knwmhosp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2168 \tab 366 \tab 355 \tab 8 \tab 7 \tab 2904 \cr
   2002 \tab 1859 \tab 561 \tab 343 \tab 2 \tab - \tab 2765 \cr
   Total \tab 4027 \tab 927 \tab 698 \tab 10 \tab 7 \tab 5669 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}}
}
\concept{Mental Health}
\keyword{variable}
