% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{knwmhosp}
\alias{knwmhosp}
\title{Know anyone who was in a hospital}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwmhosp
}
\details{
Question Did you ever know anyone who was in a hospital because of a mental illness?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2058/vshow}{GSS Data Explorer page for \code{knwmhosp}}.
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}}
}
\concept{Mental Health}
\keyword{variable}
