% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_221.R
\name{knwmw4}
\alias{knwmw4}
\title{Gender identity of person 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwmw4
}
\details{
Question Is (NAME 4), the fourth person, a man or a woman?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} transgender ftm
\item \code{4} transgender mtf
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7098/vshow}{GSS Data Explorer page for \code{knwmw4}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab male \tab female \tab other \tab don't know \tab no answer \tab Total \cr
   2018 \tab 2190 \tab 69 \tab 80 \tab 2 \tab 4 \tab 3 \tab 2348 \cr
   Total \tab 2190 \tab 69 \tab 80 \tab 2 \tab 4 \tab 3 \tab 2348 \cr
}
}

\keyword{variable}
