% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{knwpolar}
\alias{knwpolar}
\title{Level of r's information: the north and south poles}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwpolar
}
\details{
Question We want to ask about how much information you have on various topics. Please look at Card B21. For each of the following areas, please indicate whether you are very informed, somewhat informed, neither informed nor uninformed, somewhat uniformed, or very uninformed about the issues.
E. The North and South poles
}
\section{Values}{

\itemize{
\item \code{1} very informed
\item \code{2} somewhat informed
\item \code{3} neither informed nor uninformed
\item \code{4} somewhat uninformed
\item \code{5} very uninformed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3418/vshow}{GSS Data Explorer page for \code{knwpolar}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{biosci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwsci}},
\code{\link{medsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
