% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_377.R
\name{knwsci}
\alias{knwsci}
\title{Level of r's information: science and technology}
\source{
General Social Survey https://gss.norc.org
}
\description{
knwsci
}
\details{
Question We want to ask about how much information you have on various topics. Please look at Card B21. For each of the following areas, please indicate whether you are very informed, somewhat informed, neither informed nor uninformed, somewhat uniformed, or very uninformed about the issues.
C. Science and technology
}
\section{Values}{

\itemize{
\item \code{1} very informed
\item \code{2} somewhat informed
\item \code{3} neither informed nor uninformed
\item \code{4} somewhat uninformed
\item \code{5} very uninformed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3416/vshow}{GSS Data Explorer page for \code{knwsci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very informed \tab somewhat informed \tab neither informed nor uninformed \tab somewhat uninformed \tab very uninformed \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 165 \tab 909 \tab 361 \tab 245 \tab 154 \tab 25 \tab 5 \tab 4510 \cr
   Total \tab 2646 \tab 165 \tab 909 \tab 361 \tab 245 \tab 154 \tab 25 \tab 5 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{biosci}},
\code{\link{econsci}},
\code{\link{engnrsci}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{medsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
