% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{laborpow}
\alias{laborpow}
\title{Power of labor unions in r's country}
\source{
General Social Survey https://gss.norc.org
}
\description{
laborpow
}
\details{
Question Do you think that labor unions in this country have too much power or too little power?
}
\section{Values}{

\itemize{
\item \code{1} far too much power
\item \code{2} too much power
\item \code{3} about the right amount of power
\item \code{4} too little power
\item \code{5} far too little power
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4068/vshow}{GSS Data Explorer page for \code{laborpow}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
