% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{lackinfo}
\alias{lackinfo}
\title{Ppl at work fail to give r necessary info}
\source{
General Social Survey https://gss.norc.org
}
\description{
lackinfo
}
\details{
Question E. People at work fail to give me information that is necessary for me to do my job.
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3002/vshow}{GSS Data Explorer page for \code{lackinfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 139 \tab 495 \tab 516 \tab 656 \tab 20 \tab 41 \tab 2812 \cr
   2012 \tab 795 \tab 102 \tab 362 \tab 305 \tab 366 \tab 12 \tab 32 \tab 1974 \cr
   Total \tab 1740 \tab 241 \tab 857 \tab 821 \tab 1022 \tab 32 \tab 73 \tab 4786 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{difstand}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{putdown}},
\code{\link{rptprobs}}
}
\concept{Workplace Conflict}
\keyword{variable}
