% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{laidoff}
\alias{laidoff}
\title{R was laid off main job last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
laidoff
}
\details{
Question Were you laid off your main job at any time in the last year?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2818/vshow}{GSS Data Explorer page for \code{laidoff}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 131 \tab 1644 \tab 6 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 115 \tab 1602 \tab 3 \tab 14 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 112 \tab 1048 \tab 3 \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 69 \tab 1175 \tab - \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 67 \tab 1346 \tab 4 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 115 \tab 1839 \tab 7 \tab 14 \tab 14 \tab 3544 \cr
   Total \tab 8345 \tab 609 \tab 8654 \tab 23 \tab 104 \tab 14 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}},
\code{\link{wksexism}}
}
\concept{Quality of Working Life}
\keyword{variable}
