% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{lastslf}
\alias{lastslf}
\title{Current job self-employment?}
\source{
General Social Survey https://gss.norc.org
}
\description{
lastslf
}
\details{
Question And how about your job now? (If you are not working now, please tell us about your last job.)
A. Are (or were) you self-employed, or do you work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{3} never had job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4319/vshow}{GSS Data Explorer page for \code{lastslf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab self-employed \tab someone else \tab never had job \tab no answer \tab Total \cr
   1987 \tab 255 \tab 143 \tab 1176 \tab 46 \tab 199 \tab 1819 \cr
   Total \tab 255 \tab 143 \tab 1176 \tab 46 \tab 199 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
