% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{lastwkmo}
\alias{lastwkmo}
\title{What month were you last working}
\source{
General Social Survey https://gss.norc.org
}
\description{
lastwkmo
}
\details{
Question When were you last working for pay, that is what month and year were you last working?
\enumerate{
\item MONTH
}
}
\section{Values}{

\itemize{
\item \verb{[1]} january
\item \verb{[2]} february
\item \verb{[3]} march
\item \verb{[4]} april
\item \verb{[5]} may
\item \verb{[6]} june
\item \verb{[7]} july
\item \verb{[8]} august
\item \verb{[9]} september
\item \verb{[10]} october
\item \verb{[11]} november
\item \verb{[12]} december
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2868/vshow}{GSS Data Explorer page for \code{lastwkmo}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab january \tab february \tab march \tab april \tab may \tab june \tab july \tab august \tab september \tab october \tab november \tab december \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1817 \tab 105 \tab 74 \tab 79 \tab 75 \tab 72 \tab 109 \tab 66 \tab 55 \tab 68 \tab 46 \tab 40 \tab 83 \tab 70 \tab 6 \tab 2765 \cr
   Total \tab 1817 \tab 105 \tab 74 \tab 79 \tab 75 \tab 72 \tab 109 \tab 66 \tab 55 \tab 68 \tab 46 \tab 40 \tab 83 \tab 70 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{cowrkhrd}},
\code{\link{donothng}},
\code{\link{lastwkyr}},
\code{\link{talkemp}},
\code{\link{talksup}}
}
\concept{Employee Compensation}
\keyword{variable}
