% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{lastwkyr}
\alias{lastwkyr}
\title{What year were you last working}
\source{
General Social Survey https://gss.norc.org
}
\description{
lastwkyr
}
\details{
Question When were you last working for pay, that is what month and year were you last working?
2. YEAR
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2869/vshow}{GSS Data Explorer page for \code{lastwkyr}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{cowrkhrd}},
\code{\link{donothng}},
\code{\link{lastwkmo}},
\code{\link{talkemp}},
\code{\link{talksup}}

Other Work: 
\code{\link{lastwkmo}}
}
\concept{Employee Compensation}
\concept{Work}
\keyword{variable}
