% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{lastyrwk}
\alias{lastyrwk}
\title{Thinking about last full-time job, what yr r worked last}
\source{
General Social Survey https://gss.norc.org
}
\description{
lastyrwk
}
\details{
Question Thinking about your last full-time job, what year did you last work?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3893/vshow}{GSS Data Explorer page for \code{lastyrwk}}.
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
