% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_565.R
\name{law2}
\alias{law2}
\title{A personal assault}
\source{
General Social Survey https://gss.norc.org
}
\description{
law2
}
\details{
Question Now, I'm going to ask you about various events and conditions that happen to people. I'm interested in those that happened to you during the last 12 months, that is since (CURRENT MONTH), (1990/2003). As I ask you about the specific events, please think carefully, so I can record things accurately.
F. Next, did any of the following criminal or legal events occur to you since (CURRENT MONTH), (1990/2003)...
2. A personal assault (e.g. being beaten up, hit, or attacked with a weapon).
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   B/C/- \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5179/vshow}{GSS Data Explorer page for \code{law2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 500 \tab 22 \tab 991 \tab 4 \tab 1517 \cr
   2004 \tab 1472 \tab 37 \tab 1296 \tab 7 \tab 2812 \cr
   Total \tab 1972 \tab 59 \tab 2287 \tab 11 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{famper2}},
\code{\link{famper3}},
\code{\link{famper4}},
\code{\link{famper5}},
\code{\link{famper6}},
\code{\link{law1}},
\code{\link{law3}},
\code{\link{law4}},
\code{\link{law5}}
}
\concept{Negative Life Events}
\keyword{variable}
