% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{lawfirm}
\alias{lawfirm}
\title{How firm is r's opinion on law issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawfirm
}
\details{
Question How firm are you about your opinion on law enforcement--would you say you are very likely to change your opinion, somewhat likely to change, somewhat unlikely to change, or very unlikely to change?
}
\section{Values}{

\itemize{
\item \code{1} change very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} change very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/283/vshow}{GSS Data Explorer page for \code{lawfirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab change very likely \tab somewhat likely \tab somewhat unlikely \tab change very unlikely \tab don't know \tab no answer \tab Total \cr
   1984 \tab 484 \tab 28 \tab 283 \tab 313 \tab 343 \tab 20 \tab 2 \tab 1473 \cr
   Total \tab 484 \tab 28 \tab 283 \tab 313 \tab 343 \tab 20 \tab 2 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}},
\code{\link{wirtap}}
}
\concept{Core}
\keyword{variable}
