% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{lawinfo}
\alias{lawinfo}
\title{How much info does r have on law issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawinfo
}
\details{
Question How much information do you have about the law enforcement issue? Do you have all of the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \verb{[1]} all info r needs
\item \verb{[2]} most of the info
\item \verb{[3]} some info
\item \verb{[4]} very little info
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/282/vshow}{GSS Data Explorer page for \code{lawinfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all info r needs \tab most of the info \tab some info \tab very little info \tab don't know \tab no answer \tab Total \cr
   1984 \tab 484 \tab 53 \tab 218 \tab 524 \tab 187 \tab 6 \tab 1 \tab 1473 \cr
   Total \tab 484 \tab 53 \tab 218 \tab 524 \tab 187 \tab 6 \tab 1 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}},
\code{\link{wirtap}}
}
\concept{Core}
\keyword{variable}
