% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_194.R
\name{lawyer}
\alias{lawyer}
\title{Hired a lawyer in the past 10 yrs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawyer
}
\details{
Question Have you hired a lawyer for any legal work in the past ten years?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1828/vshow}{GSS Data Explorer page for \code{lawyer}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   1996 \tab 1460 \tab 650 \tab 794 \tab 2904 \cr
   Total \tab 1460 \tab 650 \tab 794 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{realtor1}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{sathome}},
\code{\link{warrntyh}}
}
\concept{Markets}
\keyword{variable}
