% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{lawyere}
\alias{lawyere}
\title{I saw their advertisement}
\source{
General Social Survey https://gss.norc.org
}
\description{
lawyere
}
\details{
Question Think about the most recent time you hired a lawyer. What made you choose this law firm or attorney? (Choose all that apply)
E. I saw their advertisement.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1833/vshow}{GSS Data Explorer page for \code{lawyere}}.
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{rellaw}},
\code{\link{satlawyr}}

Other Purchases/Transactions: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyerf}},
\code{\link{rellaw}},
\code{\link{satlawyr}}
}
\concept{Markets}
\concept{Purchases/Transactions}
\keyword{variable}
