% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_200.R
\name{layoffs}
\alias{layoffs}
\title{Who has the greatest right to keep their jobs?}
\source{
General Social Survey https://gss.norc.org
}
\description{
layoffs
}
\details{
Question From time to time companies must lay-off some of their employees. If some have to be laid-off, who, in your opinion, has the greatest right to keep their jobs?
}
\section{Values}{

\itemize{
\item \code{1} those who have worked longest
\item \code{2} those who have the greatest needs
\item \code{3} those who are the best workers
\item \code{4} difficulty in finding a new job
\item \code{5} union members
\item \code{6} older employees
\item \code{7} younger employees
\item \code{8} men
\item \code{9} women
\item \code{10} whites
\item \code{11} minorities-non-whites
\item \code{12} other -please specify
\item \code{13} owners,bosses
\item \code{14} essential employees
\item \code{15} newest employees
\item \code{16} most loyal employees
\item \code{17} production. hands-on employees
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1884/vshow}{GSS Data Explorer page for \code{layoffs}}.
}

\seealso{
Other Markets: 
\code{\link{decbed}},
\code{\link{decgift}}
}
\concept{Employment}
\concept{Job}
\concept{Markets}
\concept{Occupation}
\concept{Work}
\keyword{variable}
