% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{ldcgap}
\alias{ldcgap}
\title{Differences between rich and poor too large}
\source{
General Social Survey https://gss.norc.org
}
\description{
ldcgap
}
\details{
Question Turning to international differences, do you agree or disagree:
A. Present economic differences between rich and poor countries are too large.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4324/vshow}{GSS Data Explorer page for \code{ldcgap}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}

Other Foreign Affairs: 
\code{\link{ldctax}}
}
\concept{Foreign Affairs}
\concept{ISSP Social Inequality}
\keyword{variable}
