% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{leadgrp}
\alias{leadgrp}
\title{Has r served as officer in group}
\source{
General Social Survey https://gss.norc.org
}
\description{
leadgrp
}
\details{
Question Here is a list of things that members of organizations may participate in. Could you tell me whether or not you have ever done any of these things as a member of (GROUP IN MEMFRAT-MEMNUM OR 332)? (READ EACH ITEM. CIRCLE A NUMBER FOR EACH ACTIVITY).
B. Served as an officer
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1016/vshow}{GSS Data Explorer page for \code{leadgrp}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1987 \tab 606 \tab 541 \tab 634 \tab 1 \tab 37 \tab 1819 \cr
   Total \tab 606 \tab 541 \tab 634 \tab 1 \tab 37 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actother}},
\code{\link{attndgrp}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{givegrp}},
\code{\link{mostact}},
\code{\link{servegrp}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
