% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_570.R
\name{leakinfo}
\alias{leakinfo}
\title{Punishment for official leaking information}
\source{
General Social Survey https://gss.norc.org
}
\description{
leakinfo
}
\details{
Question Now I''m going to ask you about several offenses. For each case, indicate (Number 1-6) what you would consider the appropriate punishment for each case.
B. A high-placed government official leaked sensitive information to a newspaper in order to influence public opinion about a controversial political issue. The official received no money.
}
\section{Values}{

\itemize{
\item \code{1} reprimand
\item \code{2} dismiss military
\item \code{3} 1 year in prison
\item \code{4} 5-9 yrs prison
\item \code{5} 10-20 yrs prison
\item \code{6} life in prison
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5228/vshow}{GSS Data Explorer page for \code{leakinfo}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab reprimand \tab dismiss military \tab 1 year in prison \tab 5-9 yrs prison \tab 10-20 yrs prison \tab life in prison \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 273 \tab 591 \tab 185 \tab 156 \tab 68 \tab 25 \tab 68 \tab 21 \tab 2832 \cr
   Total \tab 1445 \tab 273 \tab 591 \tab 185 \tab 156 \tab 68 \tab 25 \tab 68 \tab 21 \tab 2832 \cr
}
}

\seealso{
Other National Security: 
\code{\link{chkfinan}},
\code{\link{chkspfin}},
\code{\link{chktaxes}},
\code{\link{knomentl}},
\code{\link{punarms}},
\code{\link{spyenemy}},
\code{\link{spyfrend}},
\code{\link{takearms}},
\code{\link{taketrck}}
}
\concept{National Security}
\keyword{variable}
