% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{learnhme}
\alias{learnhme}
\title{How learned the availability of the home}
\source{
General Social Survey https://gss.norc.org
}
\description{
learnhme
}
\details{
Question How did you learn about the availability of the home you purchased?
}
\section{Values}{

\itemize{
\item \verb{[1]} from a relative, including in-laws
\item \verb{[2]} a friend or acquaintance
\item \verb{[3]} a friend of a friend or relative, etc
\item \verb{[4]} from an advertisement or sign
\item \verb{[5]} from a real estate agent
\item \verb{[6]} i built it myself
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1814/vshow}{GSS Data Explorer page for \code{learnhme}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab from a relative, including in-laws \tab a friend or acquaintance \tab a friend of a friend or relative, etc \tab from an advertisement or sign \tab from a real estate agent \tab i built it myself \tab no answer \tab Total \cr
   1996 \tab 2451 \tab 49 \tab 63 \tab 13 \tab 116 \tab 146 \tab 57 \tab 9 \tab 2904 \cr
   Total \tab 2451 \tab 49 \tab 63 \tab 13 \tab 116 \tab 146 \tab 57 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{relhome}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
