% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{leavejb}
\alias{leavejb}
\title{Why did r leave last job}
\source{
General Social Survey https://gss.norc.org
}
\description{
leavejb
}
\details{
Question Why did you leave your last (full-time) job? Did you leave voluntarily, was the job abolished, or were you permanently laid off?
}
\section{Values}{

\itemize{
\item \verb{[1]} left voluntarily
\item \verb{[2]} job was abolished
\item \verb{[3]} permanently laid off
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3894/vshow}{GSS Data Explorer page for \code{leavejb}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab left voluntarily \tab job was abolished \tab permanently laid off \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1820 \tab 124 \tab 18 \tab 37 \tab 17 \tab 7 \tab 2023 \cr
   Total \tab 1820 \tab 124 \tab 18 \tab 37 \tab 17 \tab 7 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
