% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{leavejob}
\alias{leavejob}
\title{How likely will try to find another firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
leavejob
}
\details{
Question All in all, how likely is it that you will try to find a job with another firm or organization within the next 12 months?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} likely
\item \verb{[3]} unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4591/vshow}{GSS Data Explorer page for \code{leavejob}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab likely \tab unlikely \tab very unlikely \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 115 \tab 134 \tab 201 \tab 354 \tab 26 \tab 35 \tab 2832 \cr
   2006 \tab 3492 \tab 152 \tab 153 \tab 267 \tab 436 \tab 6 \tab 4 \tab 4510 \cr
   2016 \tab 1927 \tab 152 \tab 127 \tab 269 \tab 384 \tab 3 \tab 5 \tab 2867 \cr
   Total \tab 7386 \tab 419 \tab 414 \tab 737 \tab 1174 \tab 35 \tab 44 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
