% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{leavejob}
\alias{leavejob}
\title{How likely will try to find another firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
leavejob
}
\details{
Question All in all, how likely is it that you will try to find a job with another firm or organization within the next 12 months?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} likely
\item \code{3} unlikely
\item \code{4} very unlikely
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4591/vshow}{GSS Data Explorer page for \code{leavejob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
