% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{leftrght1}
\alias{leftrght1}
\title{How left or right in politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
leftrght1
}
\details{
Question In politics people sometimes talk of left and right. Where would you place yourself on a scale from 0 to 10 where 0 means the left and 10 means the right?
}
\section{Values}{

\itemize{
\item \verb{[0]} left
\item \verb{[10]} right
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8130/vshow}{GSS Data Explorer page for \code{leftrght1}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab not available in this year \tab left \tab 1 \tab right \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab iap \tab no answer \tab Total \cr
   2024 \tab - \tab 57 \tab 29 \tab 161 \tab 92 \tab 93 \tab 108 \tab 401 \tab 114 \tab 120 \tab 93 \tab 51 \tab 291 \tab 1665 \tab 34 \tab 3309 \cr
   Total \tab 0 \tab 57 \tab 29 \tab 161 \tab 92 \tab 93 \tab 108 \tab 401 \tab 114 \tab 120 \tab 93 \tab 51 \tab 291 \tab 1665 \tab 34 \tab 3309 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{poltalk}},
\code{\link{servpub}}
}
\concept{ISSP Citizenship}
\keyword{variable}
