% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{leftrght1}
\alias{leftrght1}
\title{How left or right in politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
leftrght1
}
\details{
Question In politics people sometimes talk of left and right. Where would you place yourself on a scale from 0 to 10 where 0 means the left and 10 means the right?
}
\section{Values}{

\itemize{
\item \code{0} left
\item \code{10} right
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8130/vshow}{GSS Data Explorer page for \code{leftrght1}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{poltalk}},
\code{\link{servpub}},
\code{\link{usmca1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
