% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_460.R
\name{lentto}
\alias{lentto}
\title{Lent money to another person past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
lentto
}
\details{
Question During the past 12 months, how often have you done any of the following things for people you know personally, such as relatives, friends, neighbors or other acquaintances?
B. Lent quite a bit of money to another person
}
\section{Values}{

\itemize{
\item \code{1} more than once a week
\item \code{2} once a week
\item \code{3} once a month
\item \code{4} at least two or three times in the past year
\item \code{5} once in the past year
\item \code{6} not at all in the past year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4221/vshow}{GSS Data Explorer page for \code{lentto}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab more than once a week \tab once a week \tab once a month \tab at least two or three times in the past year \tab once in the past year \tab not at all in the past year \tab no answer \tab don't know \tab Total \cr
   2002 \tab 1616 \tab 11 \tab 23 \tab 61 \tab 201 \tab 235 \tab 610 \tab 8 \tab - \tab 2765 \cr
   2004 \tab 1472 \tab 22 \tab 19 \tab 116 \tab 307 \tab 269 \tab 602 \tab 3 \tab 2 \tab 2812 \cr
   2012 \tab 666 \tab 11 \tab 17 \tab 56 \tab 208 \tab 258 \tab 747 \tab 7 \tab 4 \tab 1974 \cr
   2014 \tab 1264 \tab 21 \tab 15 \tab 75 \tab 220 \tab 237 \tab 701 \tab 5 \tab - \tab 2538 \cr
   Total \tab 5018 \tab 65 \tab 74 \tab 308 \tab 936 \tab 999 \tab 2660 \tab 23 \tab 6 \tab 10089 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{firstyou}},
\code{\link{frdenjoy}},
\code{\link{frdhelps}},
\code{\link{frdknows}},
\code{\link{frdthink}},
\code{\link{helpjob}},
\code{\link{kidpars}},
\code{\link{learnjob}},
\code{\link{talkedto}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
