% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{lessreg}
\alias{lessreg}
\title{Less govt regulation of business}
\source{
General Social Survey https://gss.norc.org
}
\description{
lessreg
}
\details{
Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it.
E. Less government regulation of business.
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor of
\item \code{2} in favor
\item \code{3} neither in favor nor against
\item \code{4} against
\item \code{5} strongly against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4053/vshow}{GSS Data Explorer page for \code{lessreg}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{makejobs}},
\code{\link{poleff10}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}},
\code{\link{spenviro}}

Other Economy: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{makejobs}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{setwage}}
}
\concept{Economy}
\concept{ISSP Role of Gov't}
\keyword{variable}
