% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{letdie1y}
\alias{letdie1y}
\title{Allow incurable patients to die (form 2)}
\source{
General Social Survey https://gss.norc.org
}
\description{
letdie1y
}
\details{
Question When a person has a disease that cannot be cured, do you think doctors should or should not be allowed by law to end the patient’s life by some painless means if the patient and his family request it?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2008, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5573/vshow}{GSS Data Explorer page for \code{letdie1y}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2008 \tab 1342 \tab 453 \tab 211 \tab 16 \tab 1 \tab - \tab 2023 \cr
   2021 \tab 2658 \tab 986 \tab 370 \tab 2 \tab 4 \tab 12 \tab 4032 \cr
   2022 \tab 2373 \tab 829 \tab 305 \tab 28 \tab 6 \tab 3 \tab 3544 \cr
   Total \tab 6373 \tab 2268 \tab 886 \tab 46 \tab 11 \tab 15 \tab 9599 \cr
}
}

\seealso{
Other Form Y: 
\code{\link{polhitoy}},
\code{\link{popespky}}
}
\concept{Form Y}
\keyword{variable}
