% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{letdie2}
\alias{letdie2}
\title{Let patients die if doctors agree}
\source{
General Social Survey https://gss.norc.org
}
\description{
letdie2
}
\details{
Question When a person has a disease that cannot be cured, do you think Doctors should be allowed by law to end the patient's life by some painless means if the patient and his family request it?
A. Would you approve of ending a patient's life if a board of doctors appointed by the court agreed that the patient could not be cured?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1978, 1982, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/648/vshow}{GSS Data Explorer page for \code{letdie2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1977 \tab 907 \tab 66 \tab 488 \tab 53 \tab 16 \tab 1530 \cr
   1978 \tab 884 \tab 58 \tab 533 \tab 42 \tab 15 \tab 1532 \cr
   1982 \tab 1026 \tab 46 \tab 707 \tab 59 \tab 22 \tab 1860 \cr
   1983 \tab 1005 \tab 36 \tab 496 \tab 53 \tab 9 \tab 1599 \cr
   Total \tab 3822 \tab 206 \tab 2224 \tab 207 \tab 62 \tab 6521 \cr
}
}

\seealso{
Other Core: 
\code{\link{letdie1}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}}
}
\concept{Core}
\keyword{variable}
