% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{letdie2}
\alias{letdie2}
\title{Let patients die if doctors agree}
\source{
General Social Survey https://gss.norc.org
}
\description{
letdie2
}
\details{
Question When a person has a disease that cannot be cured, do you think Doctors should be allowed by law to end the patient's life by some painless means if the patient and his family request it?
A. Would you approve of ending a patient's life if a board of doctors appointed by the court agreed that the patient could not be cured?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/648/vshow}{GSS Data Explorer page for \code{letdie2}}.
}

\seealso{
Other Core: 
\code{\link{letdie1}},
\code{\link{suicide1}},
\code{\link{suicide2}},
\code{\link{suicide3}},
\code{\link{suicide4}}
}
\concept{Core}
\keyword{variable}
