% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{letin}
\alias{letin}
\title{Increase or decrease immigrants to us}
\source{
General Social Survey https://gss.norc.org
}
\description{
letin
}
\details{
Question Do you think the number of immigrants from foreign countries who are permitted to come to the United States to live should be increased a lot, increased a little, left the same as it is now, decreased a little, or decreased a lot?
}
\section{Values}{

\itemize{
\item \verb{[1]} increased a lot
\item \verb{[2]} increased a little
\item \verb{[3]} left the same as it is now
\item \verb{[4]} decreased a little
\item \verb{[5]} decreased a lot
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1611/vshow}{GSS Data Explorer page for \code{letin}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab left the same as it is now \tab decreased a little \tab decreased a lot \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 35 \tab 53 \tab 395 \tab 407 \tab 507 \tab 69 \tab 8 \tab 2992 \cr
   2000 \tab 1419 \tab 50 \tab 74 \tab 612 \tab 281 \tab 301 \tab 76 \tab 4 \tab 2817 \cr
   Total \tab 2937 \tab 85 \tab 127 \tab 1007 \tab 688 \tab 808 \tab 145 \tab 12 \tab 5809 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{symptblk}}
}
\concept{Multiculturalism}
\keyword{variable}
