% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{letin1}
\alias{letin1}
\title{Number of immigrants to america nowadays should be}
\source{
General Social Survey https://gss.norc.org
}
\description{
letin1
}
\details{
Question Do you think the number of immigrants to America nowadays should be…
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} remain the same as it is
\item \code{4} reduced a little
\item \code{5} reduced a lot
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4863/vshow}{GSS Data Explorer page for \code{letin1}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{refugees}},
\code{\link{res161}}

Other United States: 
\code{\link{immameco}},
\code{\link{immcrime}},
\code{\link{immideas}},
\code{\link{immjobs}},
\code{\link{meltpot1}},
\code{\link{refugees}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
