% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{letin1}
\alias{letin1}
\title{Number of immigrants to america nowadays should be}
\source{
General Social Survey https://gss.norc.org
}
\description{
letin1
}
\details{
Question Do you think the number of immigrants to America nowadays should be…
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} remain the same as it is
\item \code{4} reduced a little
\item \code{5} reduced a lot
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4863/vshow}{GSS Data Explorer page for \code{letin1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab remain the same as it is \tab reduced a little \tab reduced a lot \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 33 \tab 65 \tab 310 \tab 339 \tab 394 \tab 172 \tab 54 \tab 2904 \cr
   2004 \tab 1596 \tab 45 \tab 78 \tab 354 \tab 338 \tab 278 \tab 122 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 41 \tab 114 \tab 465 \tab 270 \tab 210 \tab 170 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 119 \tab 257 \tab 1129 \tab 947 \tab 882 \tab 464 \tab 59 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{refugees}},
\code{\link{res161}}
}
\concept{ISSP National Identity}
\keyword{variable}
