% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{letin1a}
\alias{letin1a}
\title{Number of immigrants nowadays should be}
\source{
General Social Survey https://gss.norc.org
}
\description{
letin1a
}
\details{
Question Do you think the number of immigrants to America nowadays should be…
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} remain the same as it is
\item \code{4} reduced a little
\item \code{5} reduced a lot
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4901/vshow}{GSS Data Explorer page for \code{letin1a}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{shortcom}}

Other Immigration: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{patriot4}}
}
\concept{ISSP National Identity}
\concept{Immigration}
\keyword{variable}
