% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{letinasn}
\alias{letinasn}
\title{Immigration from asia should be increased}
\source{
General Social Survey https://gss.norc.org
}
\description{
letinasn
}
\details{
Question What about the number of immigrants from Asia? (Should it be increased a lot, increased a little, left the same as it is now, decreased a little, or decreased a lot?)
}
\section{Values}{

\itemize{
\item \verb{[1]} increased a lot
\item \verb{[2]} increased a little
\item \verb{[3]} left the same as it is now
\item \verb{[4]} decreased a little
\item \verb{[5]} decreased a lot
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   A/C/- \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2305/vshow}{GSS Data Explorer page for \code{letinasn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab left the same as it is now \tab decreased a little \tab decreased a lot \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 36 \tab 79 \tab 626 \tab 259 \tab 299 \tab 95 \tab 4 \tab 2817 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 4963 \tab 36 \tab 79 \tab 626 \tab 259 \tab 299 \tab 95 \tab 4 \tab 6361 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{othhome}},
\code{\link{spklangw}},
\code{\link{uswht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
