% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{letineur}
\alias{letineur}
\title{Immigration from europe should be increased}
\source{
General Social Survey https://gss.norc.org
}
\description{
letineur
}
\details{
Question What about the number of immigrants from Europe? (should it be increased a lot, increased a little, left the same as it is now, decreased a little, or decreased a lot?)
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} left the same as it is now
\item \code{4} decreased a little
\item \code{5} decreased a lot
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2306/vshow}{GSS Data Explorer page for \code{letineur}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab increased a lot \tab increased a little \tab left the same as it is now \tab decreased a little \tab decreased a lot \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 40 \tab 90 \tab 717 \tab 225 \tab 232 \tab 90 \tab 4 \tab 2817 \cr
   Total \tab 1419 \tab 40 \tab 90 \tab 717 \tab 225 \tab 232 \tab 90 \tab 4 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letinhsp}},
\code{\link{othhome}},
\code{\link{spklangw}},
\code{\link{uswht}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
