% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{letinhsp}
\alias{letinhsp}
\title{Immigration from latin america should be increased}
\source{
General Social Survey https://gss.norc.org
}
\description{
letinhsp
}
\details{
Question What about the number of immigrants from Latin America (that is, Spanish-speaking countries of the Americas)–should it be increased a lot, increased a little, left the same as it is now, decreased a little, or decreased a lot?
}
\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} left the same as it is now
\item \code{4} decreased a little
\item \code{5} decreased a lot
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2304/vshow}{GSS Data Explorer page for \code{letinhsp}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{othhome}},
\code{\link{spklangw}},
\code{\link{uswht}}

Other Immigration: 
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}}
}
\concept{Immigration}
\concept{Multi-Ethnic United States}
\keyword{variable}
