% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{letters}
\alias{letters}
\title{How many people write letters to}
\source{
General Social Survey https://gss.norc.org
}
\description{
letters
}
\details{
Question Of these____friends and relatives, about how many do you stay in contact with by:
C. Exchanging cards or letters through U.S. postal mail?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1 - 2
\item \code{3} 3 - 5
\item \code{4} 6 - 10
\item \code{5} 11 - 15
\item \code{6} 16-25
\item \code{7} 26-50
\item \code{8} 50+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2546/vshow}{GSS Data Explorer page for \code{letters}}.
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}},
\code{\link{polinfgn}}

Other Family: 
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{meetings}},
\code{\link{numcntct}}
}
\concept{Family}
\concept{Information Society}
\keyword{variable}
