% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{letters}
\alias{letters}
\title{How many people write letters to}
\source{
General Social Survey https://gss.norc.org
}
\description{
letters
}
\details{
Question Of these____friends and relatives, about how many do you stay in contact with by:
C. Exchanging cards or letters through U.S. postal mail?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1 - 2
\item \code{3} 3 - 5
\item \code{4} 6 - 10
\item \code{5} 11 - 15
\item \code{6} 16-25
\item \code{7} 26-50
\item \code{8} 50+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2546/vshow}{GSS Data Explorer page for \code{letters}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 - 2 \tab 3 - 5 \tab 6 - 10 \tab 11 - 15 \tab 16-25 \tab 26-50 \tab 50+ \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1884 \tab 227 \tab 155 \tab 186 \tab 137 \tab 44 \tab 64 \tab 48 \tab 45 \tab 1 \tab 26 \tab 2817 \cr
   2002 \tab 74 \tab 755 \tab 406 \tab 468 \tab 354 \tab 170 \tab 204 \tab 208 \tab 112 \tab 1 \tab 13 \tab 2765 \cr
   2004 \tab 1873 \tab 282 \tab 154 \tab 174 \tab 105 \tab 42 \tab 70 \tab 58 \tab 41 \tab 1 \tab 12 \tab 2812 \cr
   Total \tab 3831 \tab 1264 \tab 715 \tab 828 \tab 596 \tab 256 \tab 338 \tab 314 \tab 198 \tab 3 \tab 51 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}},
\code{\link{polinfgn}}
}
\concept{Information Society}
\keyword{variable}
