% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{lfechnce}
\alias{lfechnce}
\title{Life is a matter of chance}
\source{
General Social Survey https://gss.norc.org
}
\description{
lfechnce
}
\details{
Question I'm going to read some statements that give reasons why a person's life turns out well or poorly. (HAND CARD BZ TO RESPONDENT.)  As I read each one, tell me whether you think it is very important, important, somewhat important, or not at all important for how somebody's life turns out? E. It's just a matter of chance.
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} somewhat important
\item \code{4} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1386/vshow}{GSS Data Explorer page for \code{lfechnce}}.
}

\seealso{
Other Culture: 
\code{\link{bigband}},
\code{\link{blugrass}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{obtohelp}},
\code{\link{opoutcme}},
\code{\link{united}}

Other Success: 
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}}
}
\concept{Culture}
\concept{Success}
\keyword{variable}
