% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{lftplane}
\alias{lftplane}
\title{Key factor that enables airplane to lift}
\source{
General Social Survey https://gss.norc.org
}
\description{
lftplane
}
\details{
Question Which of the following is a key factor that enables an airplane to lift?
}
\section{Values}{

\itemize{
\item \code{1} air pressure beneath the wing is greater than that above the wing
\item \code{2} pressure within the airplane is greater than that of the outside
\item \code{3} engine power is greater than that of friction
\item \code{4} the plane's wing is lighter than air
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3473/vshow}{GSS Data Explorer page for \code{lftplane}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
