% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{libathy}
\alias{libathy}
\title{Allow anti-religionist to teach y}
\source{
General Social Survey https://gss.norc.org
}
\description{
libathy
}
\details{
Question (There are always some people whose ideas are considered bad or dangerous by other people. For instance, somebody who is against all churches and religion…)
If some people in your community suggested that a book this person wrote against churches and religion should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7435/vshow}{GSS Data Explorer page for \code{libathy}}.
}

\keyword{variable}
