% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{libcomy}
\alias{libcomy}
\title{Allow communists book in library y}
\source{
General Social Survey https://gss.norc.org
}
\description{
libcomy
}
\details{
Question (Now, we would like to ask you some questions about a self-identified Communist…)
Suppose this person wrote a book which is in your public library. Somebody in your community suggests that the book should be removed from the library. Would you favor removing it, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7440/vshow}{GSS Data Explorer page for \code{libcomy}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab remove \tab not remove \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 2675 \tab 298 \tab 1039 \tab 3 \tab 17 \tab - \tab 4032 \cr
   2022 \tab 2367 \tab 317 \tab 824 \tab 29 \tab 4 \tab 3 \tab 3544 \cr
   Total \tab 5042 \tab 615 \tab 1863 \tab 32 \tab 21 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
