% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{libhomo}
\alias{libhomo}
\title{Allow homosexuals book in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libhomo
}
\details{
Question And what about a man who admits that he is a homosexual?
C. If some people in your community suggested that a book he wrote in favor of homosexuality should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} remove
\item \verb{[2]} not remove
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/263/vshow}{GSS Data Explorer page for \code{libhomo}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab remove \tab not remove \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1973 \tab - \tab 659 \tab 804 \tab 39 \tab 2 \tab - \tab 1504 \cr
   1974 \tab - \tab 608 \tab 813 \tab 58 \tab 5 \tab - \tab 1484 \cr
   1976 \tab - \tab 614 \tab 830 \tab 53 \tab 2 \tab - \tab 1499 \cr
   1977 \tab - \tab 631 \tab 844 \tab 51 \tab 4 \tab - \tab 1530 \cr
   1980 \tab - \tab 581 \tab 853 \tab 32 \tab 2 \tab - \tab 1468 \cr
   1982 \tab - \tab 767 \tab 1005 \tab 78 \tab 10 \tab - \tab 1860 \cr
   1984 \tab - \tab 552 \tab 873 \tab 43 \tab 5 \tab - \tab 1473 \cr
   1985 \tab - \tab 636 \tab 848 \tab 45 \tab 5 \tab - \tab 1534 \cr
   1987 \tab - \tab 739 \tab 1028 \tab 46 \tab 6 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 350 \tab 589 \tab 35 \tab 3 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 337 \tab 659 \tab 32 \tab 5 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 303 \tab 586 \tab 25 \tab 3 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 282 \tab 678 \tab 29 \tab 4 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 313 \tab 713 \tab 45 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 562 \tab 1366 \tab 61 \tab 7 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 536 \tab 1325 \tab 57 \tab 5 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 479 \tab 1316 \tab 83 \tab 4 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 476 \tab 1319 \tab 61 \tab 5 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 206 \tab 694 \tab 21 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 234 \tab 653 \tab 8 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 471 \tab 1495 \tab 33 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 314 \tab 1010 \tab 26 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 273 \tab 985 \tab 15 \tab 8 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 271 \tab 1007 \tab 22 \tab 8 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 318 \tab 1370 \tab 19 \tab 7 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 322 \tab 1537 \tab 23 \tab 8 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 239 \tab 1305 \tab 25 \tab 5 \tab - \tab 2348 \cr
   2021 \tab 2714 \tab 174 \tab 1134 \tab 1 \tab 1 \tab 8 \tab 4032 \cr
   Total \tab 20052 \tab 12247 \tab 27639 \tab 1066 \tab 130 \tab 8 \tab 61142 \cr
}
}

\seealso{
Other Core: 
\code{\link{capimp}},
\code{\link{capinfo}},
\code{\link{cappun}},
\code{\link{cappun2}},
\code{\link{colhomo}},
\code{\link{colmslm}},
\code{\link{libmslm}},
\code{\link{spkhomo}},
\code{\link{spkmslm}}
}
\concept{Core}
\keyword{variable}
