% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{libsoc}
\alias{libsoc}
\title{Allow socialists book in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libsoc
}
\details{
Question Or consider a person who favored government ownership of all the railroads and all big industries.
C. If some people in your community suggested a book he wrote favoring government ownership should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
   1973 \tab - \tab full \cr
   1974 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/251/vshow}{GSS Data Explorer page for \code{libsoc}}.
}

\seealso{
Other Core: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}

Other Civil Liberties: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Civil Liberties}
\concept{Core}
\keyword{variable}
