% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{libsoc}
\alias{libsoc}
\title{Allow socialists book in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libsoc
}
\details{
Question Or consider a person who favored government ownership of all the railroads and all big industries.
C. If some people in your community suggested a book he wrote favoring government ownership should be taken out of your public library, would you favor removing this book, or not?
}
\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/251/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab no answer \tab not remove \tab remove \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 107 \tab 11 \tab 1086 \tab 409 \tab - \tab - \tab 1613 \cr
   1973 \tab 59 \tab 3 \tab 1070 \tab 372 \tab - \tab - \tab 1504 \cr
   1974 \tab 74 \tab 5 \tab 1025 \tab 380 \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab - \tab 1490 \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab - \tab - \tab 1499 \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab - \tab - \tab 1530 \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab 1532 \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab 1468 \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab - \tab - \tab 1860 \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab 1599 \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab 1473 \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab - \tab - \tab 1534 \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab 1819 \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab - \tab - \tab 1481 \tab - \tab 1481 \cr
   1989 \tab - \tab - \tab - \tab - \tab 1537 \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab - \tab - \tab 1372 \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 240 \tab 19 \tab 3181 \tab 1161 \tab 64245 \tab 6853 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
   1973 \tab - \tab full \cr
   1974 \tab - \tab full \cr
}
}

\seealso{
Other Core: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}

Other Civil Liberties: 
\code{\link{colcom}},
\code{\link{colmil}},
\code{\link{colrac}},
\code{\link{libcom}},
\code{\link{libmil}},
\code{\link{librac}},
\code{\link{spkcom}},
\code{\link{spkmil}},
\code{\link{spkrac}}
}
\concept{Civil Liberties}
\concept{Core}
\keyword{variable}
