% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_313.R
\name{lifenow}
\alias{lifenow}
\title{R's rating of life overall now from 0-10}
\source{
General Social Survey https://gss.norc.org
}
\description{
lifenow
}
\details{
Question If you were asked to rate your overall life from 0-10, where 10 represents the best possible state and 0 represents the worst possible state, how would you rate your life today?
}
\section{Values}{

\itemize{
\item \code{0} 0 - worst possible state
\item \code{10} 10 - best possible state
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7156/vshow}{GSS Data Explorer page for \code{lifenow}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 0 - worst possible state \tab 1 \tab 10 - best possible state \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab 900 \tab 4 \tab 5 \tab 131 \tab 4 \tab 12 \tab 31 \tab 123 \tab 134 \tab 355 \tab 427 \tab 187 \tab 2 \tab 33 \tab - \tab 2348 \cr
   2022 \tab 1560 \tab - \tab 5 \tab 209 \tab 3 \tab 12 \tab 32 \tab 89 \tab 191 \tab 232 \tab 512 \tab 495 \tab 174 \tab 13 \tab 17 \tab 3544 \cr
   Total \tab 2460 \tab 4 \tab 10 \tab 340 \tab 7 \tab 24 \tab 63 \tab 212 \tab 325 \tab 587 \tab 939 \tab 682 \tab 176 \tab 46 \tab 17 \tab 5892 \cr
}
}

\keyword{variable}
