% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{likeoth}
\alias{likeoth}
\title{How much do you like or dislike the person}
\source{
General Social Survey https://gss.norc.org
}
\description{
likeoth
}
\details{
Question Thinking generally now, how much do you like or dislike the person/the people who made you angry or annoyed? Do you:
}
\section{Values}{

\itemize{
\item \code{1} like him-her-them a great deal
\item \code{2} like him-her-them a little
\item \code{3} feel neutral toward him-her-them
\item \code{4} dislike him or her-them a little
\item \code{5} dislike him or her-them a lot
\item \code{6} if volunteered - love
\item \code{7} if volunteered - dont know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1786/vshow}{GSS Data Explorer page for \code{likeoth}}.
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}

Other Feelings: 
\code{\link{chnang16}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}}
}
\concept{Emotions}
\concept{Feelings}
\keyword{variable}
