% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_273.R
\name{litauth}
\alias{litauth}
\title{Visit site associated with writer}
\source{
General Social Survey https://gss.norc.org
}
\description{
litauth
}
\details{
Question D. Visited a Web site associated with a particular novelist, playwright or poet?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2482/vshow}{GSS Data Explorer page for \code{litauth}}.
}

\seealso{
Other Information Society: 
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{wwwbiz}},
\code{\link{wwwcowrk}},
\code{\link{wwwgfbf}},
\code{\link{wwwlit}},
\code{\link{wwwpersn}},
\code{\link{wwwsp}}

Other Computers And Internet: 
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{wwwbiz}},
\code{\link{wwwcowrk}},
\code{\link{wwwgfbf}},
\code{\link{wwwlit}},
\code{\link{wwwpersn}},
\code{\link{wwwsp}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
