% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{litmstxt}
\alias{litmstxt}
\title{Why litmus paper doesn't change color in mixed solution}
\source{
General Social Survey https://gss.norc.org
}
\description{
litmstxt
}
\details{
Question A solution of hydrochloric acid (HCl) in water will turn blue litmus paper red. A solution of the base sodium hydroxide (NaOH) in water will turn red litmus paper blue. If the acid and base solutions are mixed in the right proportion, the resulting solution will cause neither red nor blue litmus paper to change color. Explain why the litmus paper does not change color in the mixed solution.
}
\section{Values}{

\itemize{
\item \code{1} complete
\item \code{2} partial
\item \code{3} unsatisfactory/incorrect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3475/vshow}{GSS Data Explorer page for \code{litmstxt}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{maboygrl}},
\code{\link{salth2o}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
