% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_273.R
\name{litread}
\alias{litread}
\title{Read book after seeing author on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
litread
}
\details{
Question E. Read a book by an author whose work you first discovered on the Web?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2483/vshow}{GSS Data Explorer page for \code{litread}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2554 \tab 44 \tab 104 \tab 115 \tab - \tab 2817 \cr
   2002 \tab 2555 \tab 53 \tab 110 \tab 46 \tab 1 \tab 2765 \cr
   Total \tab 5109 \tab 97 \tab 214 \tab 161 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litsite}},
\code{\link{wwwbiz}},
\code{\link{wwwcowrk}},
\code{\link{wwwgfbf}},
\code{\link{wwwlit}},
\code{\link{wwwpersn}},
\code{\link{wwwsp}}
}
\concept{Information Society}
\keyword{variable}
