% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_130.R
\name{liveblks}
\alias{liveblks}
\title{Neighborhood half black}
\source{
General Social Survey https://gss.norc.org
}
\description{
liveblks
}
\details{
Question Now I'm going to ask you about different types of contact with various groups of people. In each situation would you please tell me whether you would be very much in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very much opposed to it happening?
B. Living in a neighborhood where half of your neighbors were blacks?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/- \tab partial \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1215/vshow}{GSS Data Explorer page for \code{liveblks}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{liveasns}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrso}}

Other Blacks: 
\code{\link{patrblks}}
}
\concept{Blacks}
\concept{Intergroup Relations}
\keyword{variable}
