% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{livehome}
\alias{livehome}
\title{Length at present dwelling}
\source{
General Social Survey https://gss.norc.org
}
\description{
livehome
}
\details{
Question How long have you lived in your present home, that is, the address where you live now? PLEASE FILL IN THE NUMBER OF YEARS IN THE SPACE BELOW. IF LESS THAN ONE YEAR,  MAKE  A CHECK ON THE LINE NEXT TO THAT CHOICE.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4182/vshow}{GSS Data Explorer page for \code{livehome}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
