% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_130.R
\name{livejews}
\alias{livejews}
\title{Neighborhood half jewish}
\source{
General Social Survey https://gss.norc.org
}
\description{
livejews
}
\details{
Question Now I'm going to ask you about different types of contact with various groups of people. In each situation would you please tell me whether you would be very much in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very much opposed to it happening?
A. Living in a neighborhood where half of your neighbors were Jews?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1214/vshow}{GSS Data Explorer page for \code{livejews}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 84 \tab 237 \tab 836 \tab 148 \tab 38 \tab 19 \tab 10 \tab 1372 \cr
   2000 \tab 1419 \tab 174 \tab 297 \tab 713 \tab 96 \tab 27 \tab 77 \tab 14 \tab 2817 \cr
   Total \tab 1419 \tab 258 \tab 534 \tab 1549 \tab 244 \tab 65 \tab 96 \tab 24 \tab 4189 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrso}}
}
\concept{Intergroup Relations}
\keyword{variable}
