% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_130.R
\name{liveno}
\alias{liveno}
\title{Neighborhood half northern white}
\source{
General Social Survey https://gss.norc.org
}
\description{
liveno
}
\details{
Question Now I'm going to ask you about different types of contact with various groups of people. In each situation would you please tell me whether you would be very much in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very much opposed to it happening?
E. Living in a neighborhood where half of your neighbors were whites raised in the North?
(INTERVIEWER READ OPPOSITE REGION - SOUTH = CENSUS REGIONS - SOUTH ATLANTIC, SOUTH EAST CENTRAL, SOUTH WEST CENTRAL; NORTH = EVERYWHERE ELSE)
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1218/vshow}{GSS Data Explorer page for \code{liveno}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab 914 \tab 38 \tab 93 \tab 261 \tab 42 \tab 8 \tab 10 \tab 6 \tab 1372 \cr
   Total \tab 914 \tab 38 \tab 93 \tab 261 \tab 42 \tab 8 \tab 10 \tab 6 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveso}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}},
\code{\link{patrso}}
}
\concept{Intergroup Relations}
\keyword{variable}
