% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{livewith}
\alias{livewith}
\title{Do you live together?}
\source{
General Social Survey https://gss.norc.org
}
\description{
livewith
}
\details{
Question Do you live together?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} has same gender partner
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1650/vshow}{GSS Data Explorer page for \code{livewith}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2553 \tab 97 \tab 223 \tab 31 \tab 2904 \cr
   1998 \tab 2527 \tab 83 \tab 188 \tab 34 \tab 2832 \cr
   Total \tab 5080 \tab 180 \tab 411 \tab 65 \tab 5736 \cr
}
}

\seealso{
Other Gender: 
\code{\link{haprom}},
\code{\link{romance}}
}
\concept{Gender}
\keyword{variable}
