% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{livewith}
\alias{livewith}
\title{Do you live together?}
\source{
General Social Survey https://gss.norc.org
}
\description{
livewith
}
\details{
Question Do you live together?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} has same gender partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1650/vshow}{GSS Data Explorer page for \code{livewith}}.
}

\seealso{
Other Gender: 
\code{\link{haprom}},
\code{\link{romance}}

Other Romantic Relationships: 
\code{\link{haprom}},
\code{\link{romance}}
}
\concept{Cohabitation}
\concept{Gender}
\concept{Romantic Relationships}
\keyword{variable}
