% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{livnowed}
\alias{livnowed}
\title{Did r live with partner before marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
livnowed
}
\details{
Question Did you ever live with a partner you didn't marry?
}
\section{Values}{

\itemize{
\item \code{1} yes, with a previous partner
\item \code{2} yes, with my present partner
\item \code{3} yes, with previous and present partner
\item \code{4} no, never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4411/vshow}{GSS Data Explorer page for \code{livnowed}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{Cohabitation}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
