% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{livnowed}
\alias{livnowed}
\title{Did r live with partner before marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
livnowed
}
\details{
Question Did you ever live with a partner you didn't marry?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, with a previous partner
\item \verb{[2]} yes, with my present partner
\item \verb{[3]} yes, with previous and present partner
\item \verb{[4]} no, never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4411/vshow}{GSS Data Explorer page for \code{livnowed}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, with a previous partner \tab yes, with my present partner \tab yes, with previous and present partner \tab no, never \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 215 \tab 53 \tab 41 \tab 1001 \tab 137 \tab 2992 \cr
   Total \tab 1545 \tab 215 \tab 53 \tab 41 \tab 1001 \tab 137 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
