% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{lkelyvot}
\alias{lkelyvot}
\title{Likelihood to vote in 2024 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
lkelyvot
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} extremely likely
\item \code{2} very likely
\item \code{3} somewhat likely
\item \code{4} not very likely
\item \code{5} not at all likely
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
