% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{lngthinv}
\alias{lngthinv}
\title{How long was interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
lngthinv
}
\details{
Question How long was interview
}
\section{Values}{

\itemize{
\item \code{30} 30 minutes or less
\item \code{200} 200 minutes and more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1998, 2000, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{mode}()}
}
\concept{Paradata}
\keyword{variable}
