% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{lngthinv}
\alias{lngthinv}
\title{How long was interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
lngthinv
}
\details{
Question How long was interview
}
\section{Values}{

\itemize{
\item \verb{[30]} 30 minutes or less
\item \verb{[200]} 200 minutes and more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1998, 2000, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{mode}()}
}
\concept{Paradata}
\keyword{variable}
