% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_197.R
\name{loanedto}
\alias{loanedto}
\title{Oto whom did you lend this money?}
\source{
General Social Survey https://gss.norc.org
}
\description{
loanedto
}
\details{
Question Think of the most recent loan of this type. To whom did you lend this money?
}
\section{Values}{

\itemize{
\item \code{1} parents of grandparents
\item \code{2} children
\item \code{3} other relative
\item \code{4} friend
\item \code{5} co-worker or bus. associate
\item \code{6} other specify
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1856/vshow}{GSS Data Explorer page for \code{loanedto}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab parents of grandparents \tab children \tab other relative \tab friend \tab co-worker or bus. associate \tab other specify \tab Total \cr
   1996 \tab 2620 \tab 15 \tab 66 \tab 68 \tab 103 \tab 26 \tab 6 \tab 2904 \cr
   Total \tab 2620 \tab 15 \tab 66 \tab 68 \tab 103 \tab 26 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{loaned}},
\code{\link{sellauto}},
\code{\link{sellhome}},
\code{\link{transoth}}
}
\concept{Markets}
\keyword{variable}
