% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{loanitem}
\alias{loanitem}
\title{R has let someone borrow a item of some value}
\source{
General Social Survey https://gss.norc.org
}
\description{
loanitem
}
\details{
Question During the past 12 months, how often have you done each of the following things:
K. Let someone you didn''t know well borrow a item of some value like dishes or tools.
}
\section{Values}{

\itemize{
\item \code{1} more than once a week
\item \code{2} once a week
\item \code{3} once a month
\item \code{4} at least 2 or 3 times in the past year
\item \code{5} once in the past year
\item \code{6} not at all in the past year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2887/vshow}{GSS Data Explorer page for \code{loanitem}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab more than once a week \tab once a week \tab once a month \tab at least 2 or 3 times in the past year \tab once in the past year \tab not at all in the past year \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 11 \tab 20 \tab 49 \tab 221 \tab 231 \tab 820 \tab 14 \tab 6 \tab 2765 \cr
   2004 \tab 1472 \tab 18 \tab 19 \tab 79 \tab 253 \tab 233 \tab 733 \tab 2 \tab 3 \tab 2812 \cr
   2012 \tab 666 \tab 13 \tab 21 \tab 45 \tab 195 \tab 188 \tab 834 \tab 4 \tab 8 \tab 1974 \cr
   2014 \tab 1264 \tab 13 \tab 16 \tab 77 \tab 214 \tab 165 \tab 785 \tab - \tab 4 \tab 2538 \cr
   Total \tab 4795 \tab 55 \tab 76 \tab 250 \tab 883 \tab 817 \tab 3172 \tab 20 \tab 21 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selffrst}},
\code{\link{selfless}}
}
\concept{Altruism}
\keyword{variable}
