% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{localgvt}
\alias{localgvt}
\title{How likely can r do sth for improvmnt in community}
\source{
General Social Survey https://gss.norc.org
}
\description{
localgvt
}
\details{
Question Suppose you wanted the local government to bring about some improvement in your local community. How likely is it that you would be able to do something about it?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4240/vshow}{GSS Data Explorer page for \code{localgvt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 101 \tab 364 \tab 409 \tab 186 \tab 78 \tab 11 \tab 2765 \cr
   Total \tab 1616 \tab 101 \tab 364 \tab 409 \tab 186 \tab 78 \tab 11 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{demands}},
\code{\link{helpfrds}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{usefrds}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
