% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_267.R
\name{locate12}
\alias{locate12}
\title{R use web to find addresses in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
locate12
}
\details{
Question In the past 12 months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/02 have you used the Web to do each of these things:
I. Locate someone's phone number, e-mail, or mailing address?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2423/vshow}{GSS Data Explorer page for \code{locate12}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 370 \tab 296 \tab 76 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 549 \tab 440 \tab 94 \tab 1 \tab 2765 \cr
   Total \tab 3756 \tab 919 \tab 736 \tab 170 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{chat12}},
\code{\link{econ12}},
\code{\link{game12}},
\code{\link{health12}},
\code{\link{newjob12}},
\code{\link{people12}},
\code{\link{polinf12}},
\code{\link{travel12}}
}
\concept{Information Society}
\keyword{variable}
