% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{locgrp}
\alias{locgrp}
\title{Has r helped to start new local group}
\source{
General Social Survey https://gss.norc.org
}
\description{
locgrp
}
\details{
Question Have you ever taken part in forming a new group or a new organization to try to solve some community problems?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1025/vshow}{GSS Data Explorer page for \code{locgrp}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1987 \tab - \tab 309 \tab 1498 \tab 12 \tab 1819 \cr
   Total \tab 0 \tab 309 \tab 1498 \tab 12 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
