% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{locinflu}
\alias{locinflu}
\title{How much influence r feels on local gov}
\source{
General Social Survey https://gss.norc.org
}
\description{
locinflu
}
\details{
Question How much influence do you think people like you can have over local government decisions -- a lot, a moderate amount, a little, or none at all?
}
\section{Values}{

\itemize{
\item \code{1} a lot
\item \code{2} moderate amount
\item \code{3} a little
\item \code{4} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1023/vshow}{GSS Data Explorer page for \code{locinflu}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}

Other Political: 
\code{\link{intpol}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Political}
\concept{Socio-political participation}
\keyword{variable}
