% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{lockedup}
\alias{lockedup}
\title{Prison or jail ever}
\source{
General Social Survey https://gss.norc.org
}
\description{
lockedup
}
\details{
Question C. Have you ever spent any time in prison or jail?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/687/vshow}{GSS Data Explorer page for \code{lockedup}}.
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}},
\code{\link{tvhours}}

Other Courts: 
\code{\link{convictd}}

Other Police: 
\code{\link{arrest}},
\code{\link{ticket}}
}
\concept{Core}
\concept{Courts}
\concept{Police}
\keyword{variable}
