% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{loclived}
\alias{loclived}
\title{How long has r lived in community}
\source{
General Social Survey https://gss.norc.org
}
\description{
loclived
}
\details{
Question How long have you been living here in this community of (MENTION NAME OF LOCAL GOVERNMENT UNIT)?
}
\section{Values}{

\itemize{
\item \code{1} lt 1 year
\item \code{2} 1-3 years
\item \code{3} 4-10 years
\item \code{4} gt 10 but lt life
\item \code{5} entire life
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1021/vshow}{GSS Data Explorer page for \code{loclived}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab lt 1 year \tab 1-3 years \tab 4-10 years \tab gt 10 but lt life \tab entire life \tab no answer \tab Total \cr
   1987 \tab - \tab 172 \tab 268 \tab 314 \tab 718 \tab 339 \tab 8 \tab 1819 \cr
   Total \tab 0 \tab 172 \tab 268 \tab 314 \tab 718 \tab 339 \tab 8 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
