% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{locvote}
\alias{locvote}
\title{How often does r vote in local elections}
\source{
General Social Survey https://gss.norc.org
}
\description{
locvote
}
\details{
Question What about local elections -- do you always vote in those, do you sometimes miss one, or do you rarely vote, or do you never vote?
}
\section{Values}{

\itemize{
\item \verb{[1]} vote in all
\item \verb{[2]} sometimes miss
\item \verb{[3]} rarely vote
\item \verb{[4]} never vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1039/vshow}{GSS Data Explorer page for \code{locvote}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab vote in all \tab sometimes miss \tab rarely vote \tab never vote \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 611 \tab 598 \tab 223 \tab 363 \tab 7 \tab 17 \tab 1819 \cr
   Total \tab 0 \tab 611 \tab 598 \tab 223 \tab 363 \tab 7 \tab 17 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
