% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{losejb12}
\alias{losejb12}
\title{How likely r will lose job or laid off in next 1 yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
losejb12
}
\details{
Question Thinking now about the next 12 months, how likely do you think it is that you will lose your job or be laid off for reasons other than for your performance? Would you say…
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} fairly likely
\item \code{3} not too likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3879/vshow}{GSS Data Explorer page for \code{losejb12}}.
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}

Other Employment: 
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}}

Other Job: 
\code{\link{findnwjb}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{wkcomptr}}

Other Unemployment: 
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}}
}
\concept{Employment}
\concept{Global Economics}
\concept{Job}
\concept{Unemployment}
\keyword{variable}
