% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{losejob5}
\alias{losejob5}
\title{In last 5 yrs, r has been denied a job or fired bc employer believed r was gay/l}
\source{
General Social Survey https://gss.norc.org
}
\description{
losejob5
}
\details{
Question In the last five years, have you been denied a job or been fired from a job because an employer believed you were gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} not employed in last 5 years
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3921/vshow}{GSS Data Explorer page for \code{losejob5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab not employed in last 5 years \tab no answer \tab Total \cr
   2008 \tab 1980 \tab 5 \tab 5 \tab 1 \tab 32 \tab 2023 \cr
   2012 \tab 1921 \tab 5 \tab 1 \tab 1 \tab 46 \tab 1974 \cr
   Total \tab 3901 \tab 10 \tab 6 \tab 2 \tab 78 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{negjob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
