% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{lostjobs}
\alias{lostjobs}
\title{Jobs taken away are better than r's current job}
\source{
General Social Survey https://gss.norc.org
}
\description{
lostjobs
}
\details{
Question On average, do you think the jobs in the United States that have been taken away from people with jobs like yours are better than your job (with higher pay and better benefits) or worse?
}
\section{Values}{

\itemize{
\item \code{1} better
\item \code{2} about the same
\item \code{3} worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3886/vshow}{GSS Data Explorer page for \code{lostjobs}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab better \tab about the same \tab worse \tab don't know \tab Total \cr
   2008 \tab 1747 \tab 40 \tab 48 \tab 177 \tab 11 \tab 2023 \cr
   Total \tab 1747 \tab 40 \tab 48 \tab 177 \tab 11 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\keyword{variable}
