% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_230.R
\name{lotr1}
\alias{lotr1}
\title{In uncertain times i usually expect best}
\source{
General Social Survey https://gss.norc.org
}
\description{
lotr1
}
\details{
Question (HAND CARD G1)                       
Please answer the following questions about yourself by indicating the extent of your agreement using the following scale: Strongly disagree, Disagree, Neutral, Agree, and Strongly agree.
Be as honest as you can throughout, and try not to let your responses to one question influence your response to other questions. There are no right or wrong answers.  Do you agree or disagree that...
In uncertain times, I usually expect the best.
}
\section{Values}{

\itemize{
\item \code{1} strongly disagree
\item \code{2} disagree
\item \code{3} neutral
\item \code{4} agree
\item \code{5} strongly agree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6197/vshow}{GSS Data Explorer page for \code{lotr1}}.
}

\seealso{
Other Hope and Optimism: 
\code{\link{hope1}},
\code{\link{hope2}},
\code{\link{hope3}},
\code{\link{hope4}},
\code{\link{lotr2}},
\code{\link{lotr3}},
\code{\link{lotr4}},
\code{\link{lotr5}},
\code{\link{lotr6}}
}
\concept{Hope and Optimism}
\keyword{variable}
