% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{macall}
\alias{macall}
\title{Contact with mother}
\source{
General Social Survey https://gss.norc.org
}
\description{
macall
}
\details{
Question Is your mother still alive?
C. And how often do you have any other contact with your mother besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} several x week
\item \code{3} once a week
\item \code{4} once a month
\item \code{5} several x year
\item \code{6} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4130/vshow}{GSS Data Explorer page for \code{macall}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab no answer \tab Total \cr
   1986 \tab 678 \tab 117 \tab 187 \tab 222 \tab 168 \tab 37 \tab 54 \tab 7 \tab 1470 \cr
   Total \tab 678 \tab 117 \tab 187 \tab 222 \tab 168 \tab 37 \tab 54 \tab 7 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{malive}},
\code{\link{matime}},
\code{\link{mavisit}},
\code{\link{palive}},
\code{\link{peocntct}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
