% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{madat22}
\alias{madat22}
\title{Angry at an object, circumstances, etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
madat22
}
\details{
Question Who were you angry, irritated or annoyed at? (IF RESPONDENT GIVES A NAME, PROBE WITH "and who is name... what relationship is he/she to you?" IF UNSURE ABOUT RELATIONSHIP AFTER RESPONDENT'S DESCRIPTION, READ THE FIELD CODE THAT YOU THINK IS CLOSEST -- e.g., "so that's a co-worker?") CIRCLE ALL THAT APPLY.
V. Not angry at anyone (not angry at a person, but an object, circumstances, etc.)
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1762/vshow}{GSS Data Explorer page for \code{madat22}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1779 \tab 93 \tab 1025 \tab 7 \tab 2904 \cr
   Total \tab 1779 \tab 93 \tab 1025 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{howangry}},
\code{\link{madat23}},
\code{\link{thnkangr}}
}
\concept{Emotions}
\keyword{variable}
