% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{madat23}
\alias{madat23}
\title{Angry at other}
\source{
General Social Survey https://gss.norc.org
}
\description{
madat23
}
\details{
Question Who were you angry, irritated or annoyed at? (IF RESPONDENT GIVES A NAME, PROBE WITH "and who is name... what relationship is he/she to you?" IF UNSURE ABOUT RELATIONSHIP AFTER RESPONDENT'S DESCRIPTION, READ THE FIELD CODE THAT YOU THINK IS CLOSEST -- e.g., "so that's a co-worker?") CIRCLE ALL THAT APPLY.
W. Other (SPECIFY)
}
\section{Values}{

\itemize{
\item \code{1} ex-spouse
\item \code{2} other ex-family
\item \code{3} health care personnel
\item \code{4} teacher
\item \code{5} landlord
\item \code{6} company
\item \code{7} criminal
\item \code{8} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1763/vshow}{GSS Data Explorer page for \code{madat23}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab ex-spouse \tab other ex-family \tab health care personnel \tab teacher \tab landlord \tab company \tab criminal \tab other \tab no answer \tab Total \cr
   1996 \tab 2766 \tab 16 \tab 2 \tab 10 \tab 9 \tab 5 \tab 6 \tab 4 \tab 79 \tab 7 \tab 2904 \cr
   Total \tab 2766 \tab 16 \tab 2 \tab 10 \tab 9 \tab 5 \tab 6 \tab 4 \tab 79 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{howangry}},
\code{\link{madat22}},
\code{\link{thnkangr}}
}
\concept{Emotions}
\keyword{variable}
