% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_118.R
\name{madatgod}
\alias{madatgod}
\title{How often r feels angry at god}
\source{
General Social Survey https://gss.norc.org
}
\description{
madatgod
}
\details{
Question Here is another card with contrasting ideas. If you feel that
"My faith is completely free of doubts," you would place yourself at 1. If you feel "My faith is mixed with doubts," you would place yourself at 7. If you feel somewhere between these two, you would place yourself at 2, 3, 4, 5 or 6. Where would you place your feelings about your faith…
B. From Often felt agnry to Never felt angry at God
}
\section{Values}{

\itemize{
\item \code{1} often angry at god
\item \code{7} never angry at god
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1094/vshow}{GSS Data Explorer page for \code{madatgod}}.
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{follow}},
\code{\link{goownway}},
\code{\link{myfaith}}
}
\concept{Religion}
\keyword{variable}
