% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{maden}
\alias{maden}
\title{What was mother's religious denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
maden
}
\details{
Question A. What specific denomination is that, if any? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \code{10} am baptist asso
\item \code{11} am bapt ch in usa
\item \code{12} nat bapt conv of am
\item \code{13} nat bapt conv usa
\item \code{14} southern baptist
\item \code{15} other baptists
\item \code{18} baptist-dk which
\item \code{20} afr meth episcopal
\item \code{21} afr meth ep zion
\item \code{22} united methodist
\item \code{23} other methodist
\item \code{28} methodist-dk which
\item \code{30} am lutheran
\item \code{31} luth ch in america
\item \code{32} lutheran-mo synod
\item \code{33} wi evan luth synod
\item \code{34} other lutheran
\item \code{35} evangelical luth
\item \code{38} lutheran-dk which
\item \code{40} presbyterian c in us
\item \code{41} united pres ch in us
\item \code{42} other presbyterian
\item \code{48} presbyterian-dk wh
\item \code{50} episcopal
\item \code{60} other
\item \code{70} no denomination
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1059/vshow}{GSS Data Explorer page for \code{maden}}.
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{blkwhite}},
\code{\link{bwfriend}},
\code{\link{cathsch}},
\code{\link{chrissch}},
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{churhsch}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denomsch}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{homesch}},
\code{\link{implives}},
\code{\link{join16}},
\code{\link{kid5up}},
\code{\link{lapsed}},
\code{\link{madatgod}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{marelig}},
\code{\link{myfaith}},
\code{\link{numbwfrd}},
\code{\link{numcong}},
\code{\link{numfrend}},
\code{\link{obeytch}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{permoral}},
\code{\link{preach}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relsch}},
\code{\link{rotapple}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{tvrelig}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
