% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{maeduc}
\alias{maeduc}
\title{Highest year school completed, mother (or youngest same-sex parent)}
\source{
General Social Survey https://gss.norc.org
}
\description{
maeduc
}
\details{
Question RESPONDENT'S MOTHER'S (SUBSTITUTE MOTHER'S) EDUCATION
}
\section{Values}{

\itemize{
\item \code{0} no formal schooling
\item \code{1} 1st grade
\item \code{2} 2nd grade
\item \code{3} 3rd grade
\item \code{4} 4th grade
\item \code{5} 5th grade
\item \code{6} 6th grade
\item \code{7} 7th grade
\item \code{8} 8th grade
\item \code{9} 9th grade
\item \code{10} 10th grade
\item \code{11} 11th grade
\item \code{12} 12th grade
\item \code{13} 1 year of college
\item \code{14} 2 years of college
\item \code{15} 3 years of college
\item \code{16} 4 years of college
\item \code{17} 5 years of college
\item \code{18} 6 years of college
\item \code{19} 7 years of college
\item \code{20} 8 or more years of college
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/57/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllllllllllllll}{
   year \tab 1 year of college \tab 10th grade \tab 11th grade \tab 12th grade \tab 1st grade \tab 2 years of college \tab 2nd grade \tab 3 years of college \tab 3rd grade \tab 4 years of college \tab 4th grade \tab 5 years of college \tab 5th grade \tab 6 years of college \tab 6th grade \tab 7 years of college \tab 7th grade \tab 8th grade \tab 9th grade \tab don't know \tab iap \tab no answer \tab no formal schooling \tab 8 or more years of college \tab not available in this release \tab Total \cr
   1972 \tab 19 \tab 85 \tab 54 \tab 328 \tab 8 \tab 60 \tab 12 \tab 14 \tab 43 \tab 42 \tab 43 \tab 3 \tab 43 \tab 7 \tab 78 \tab 3 \tab 47 \tab 262 \tab 36 \tab 263 \tab 80 \tab 16 \tab 67 \tab - \tab - \tab 1613 \cr
   1973 \tab 28 \tab 74 \tab 60 \tab 342 \tab 7 \tab 50 \tab 7 \tab 8 \tab 32 \tab 47 \tab 38 \tab 2 \tab 41 \tab 4 \tab 76 \tab - \tab 39 \tab 254 \tab 40 \tab 207 \tab 67 \tab 22 \tab 56 \tab 3 \tab - \tab 1504 \cr
   1974 \tab 33 \tab 90 \tab 61 \tab 331 \tab 2 \tab 53 \tab 10 \tab 10 \tab 25 \tab 48 \tab 38 \tab 3 \tab 48 \tab 10 \tab 78 \tab 2 \tab 35 \tab 243 \tab 47 \tab 201 \tab 55 \tab 1 \tab 58 \tab 2 \tab - \tab 1484 \cr
   1975 \tab 33 \tab 90 \tab 57 \tab 360 \tab 4 \tab 49 \tab 7 \tab 7 \tab 30 \tab 53 \tab 34 \tab 1 \tab 40 \tab 8 \tab 64 \tab 2 \tab 71 \tab 253 \tab 53 \tab 193 \tab 38 \tab 3 \tab 40 \tab - \tab - \tab 1490 \cr
   1976 \tab 25 \tab 66 \tab 67 \tab 354 \tab 4 \tab 46 \tab 12 \tab 11 \tab 35 \tab 47 \tab 35 \tab 2 \tab 27 \tab 2 \tab 80 \tab 1 \tab 54 \tab 243 \tab 41 \tab 236 \tab 51 \tab 6 \tab 50 \tab 4 \tab - \tab 1499 \cr
   1977 \tab 23 \tab 78 \tab 59 \tab 378 \tab 1 \tab 44 \tab 1 \tab 10 \tab 21 \tab 50 \tab 49 \tab 3 \tab 27 \tab 9 \tab 63 \tab 1 \tab 57 \tab 256 \tab 44 \tab 243 \tab 53 \tab 4 \tab 56 \tab - \tab - \tab 1530 \cr
   1978 \tab 22 \tab 71 \tab 67 \tab 421 \tab 1 \tab 65 \tab 8 \tab 16 \tab 18 \tab 48 \tab 31 \tab 5 \tab 26 \tab 7 \tab 74 \tab 1 \tab 42 \tab 242 \tab 49 \tab 216 \tab 51 \tab 5 \tab 41 \tab 5 \tab - \tab 1532 \cr
   1980 \tab 34 \tab 67 \tab 46 \tab 383 \tab 5 \tab 69 \tab 4 \tab 5 \tab 27 \tab 62 \tab 41 \tab 7 \tab 29 \tab 9 \tab 75 \tab - \tab 39 \tab 227 \tab 50 \tab 181 \tab 62 \tab 3 \tab 41 \tab 2 \tab - \tab 1468 \cr
   1982 \tab 39 \tab 96 \tab 60 \tab 491 \tab 4 \tab 59 \tab 11 \tab 22 \tab 30 \tab 79 \tab 23 \tab 8 \tab 49 \tab 10 \tab 88 \tab - \tab 42 \tab 229 \tab 62 \tab 313 \tab 66 \tab 13 \tab 62 \tab 4 \tab - \tab 1860 \cr
   1983 \tab 40 \tab 70 \tab 68 \tab 483 \tab 2 \tab 73 \tab 13 \tab 14 \tab 24 \tab 76 \tab 33 \tab 8 \tab 33 \tab 9 \tab 61 \tab 4 \tab 46 \tab 221 \tab 46 \tab 179 \tab 48 \tab 3 \tab 44 \tab 1 \tab - \tab 1599 \cr
   1984 \tab 35 \tab 103 \tab 50 \tab 450 \tab 1 \tab 68 \tab 7 \tab 21 \tab 19 \tab 57 \tab 21 \tab 6 \tab 28 \tab 7 \tab 43 \tab 2 \tab 32 \tab 201 \tab 55 \tab 161 \tab 68 \tab 4 \tab 33 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 36 \tab 74 \tab 59 \tab 482 \tab - \tab 62 \tab 12 \tab 11 \tab 25 \tab 73 \tab 26 \tab 7 \tab 27 \tab 18 \tab 60 \tab 6 \tab 35 \tab 201 \tab 48 \tab 172 \tab 67 \tab - \tab 31 \tab 2 \tab - \tab 1534 \cr
   1986 \tab 25 \tab 59 \tab 48 \tab 472 \tab - \tab 75 \tab 5 \tab 10 \tab 24 \tab 71 \tab 22 \tab 7 \tab 29 \tab 14 \tab 50 \tab 2 \tab 40 \tab 193 \tab 48 \tab 191 \tab 47 \tab 4 \tab 33 \tab 1 \tab - \tab 1470 \cr
   1987 \tab 51 \tab 92 \tab 73 \tab 508 \tab 2 \tab 103 \tab 9 \tab 17 \tab 29 \tab 87 \tab 29 \tab 9 \tab 35 \tab 25 \tab 76 \tab 1 \tab 33 \tab 238 \tab 71 \tab 205 \tab 61 \tab 13 \tab 42 \tab 10 \tab - \tab 1819 \cr
   1988 \tab 56 \tab 66 \tab 48 \tab 470 \tab 2 \tab 79 \tab 6 \tab 15 \tab 28 \tab 74 \tab 28 \tab 4 \tab 23 \tab 8 \tab 48 \tab 3 \tab 30 \tab 176 \tab 43 \tab 177 \tab 37 \tab 18 \tab 36 \tab 6 \tab - \tab 1481 \cr
   1989 \tab 47 \tab 87 \tab 36 \tab 475 \tab 2 \tab 78 \tab 7 \tab 26 \tab 23 \tab 92 \tab 19 \tab 4 \tab 28 \tab 20 \tab 63 \tab 2 \tab 29 \tab 182 \tab 49 \tab 162 \tab 41 \tab 2 \tab 60 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 54 \tab 54 \tab 50 \tab 436 \tab 1 \tab 71 \tab 3 \tab 22 \tab 24 \tab 80 \tab 16 \tab 6 \tab 26 \tab 14 \tab 38 \tab 1 \tab 19 \tab 160 \tab 52 \tab 151 \tab 32 \tab 33 \tab 25 \tab 4 \tab - \tab 1372 \cr
   1991 \tab 56 \tab 71 \tab 43 \tab 455 \tab 1 \tab 84 \tab 7 \tab 16 \tab 15 \tab 80 \tab 14 \tab 8 \tab 24 \tab 18 \tab 39 \tab - \tab 32 \tab 192 \tab 49 \tab 176 \tab 45 \tab 63 \tab 24 \tab 5 \tab - \tab 1517 \cr
   1993 \tab 47 \tab 76 \tab 47 \tab 573 \tab 1 \tab 101 \tab 8 \tab 22 \tab 14 \tab 102 \tab 23 \tab 5 \tab 23 \tab 17 \tab 40 \tab 3 \tab 37 \tab 177 \tab 35 \tab 159 \tab 41 \tab 27 \tab 23 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 95 \tab 141 \tab 100 \tab 1007 \tab 3 \tab 197 \tab 16 \tab 47 \tab 29 \tab 187 \tab 38 \tab 15 \tab 33 \tab 38 \tab 87 \tab 7 \tab 58 \tab 319 \tab 86 \tab 284 \tab 135 \tab 13 \tab 41 \tab 16 \tab - \tab 2992 \cr
   1996 \tab 76 \tab 130 \tab 93 \tab 1009 \tab 3 \tab 243 \tab 8 \tab 46 \tab 23 \tab 214 \tab 31 \tab 16 \tab 31 \tab 48 \tab 69 \tab 8 \tab 48 \tab 243 \tab 75 \tab 292 \tab 123 \tab 17 \tab 34 \tab 24 \tab - \tab 2904 \cr
   1998 \tab 98 \tab 129 \tab 103 \tab 932 \tab 3 \tab 204 \tab 16 \tab 44 \tab 26 \tab 235 \tab 27 \tab 22 \tab 33 \tab 47 \tab 87 \tab 5 \tab 36 \tab 229 \tab 72 \tab 269 \tab 124 \tab 40 \tab 32 \tab 19 \tab - \tab 2832 \cr
   2000 \tab 88 \tab 129 \tab 81 \tab 939 \tab - \tab 201 \tab 17 \tab 41 \tab 39 \tab 218 \tab 34 \tab 17 \tab 23 \tab 41 \tab 79 \tab 4 \tab 37 \tab 233 \tab 77 \tab 284 \tab 132 \tab 33 \tab 46 \tab 24 \tab - \tab 2817 \cr
   2002 \tab 116 \tab 121 \tab 84 \tab 997 \tab 3 \tab 205 \tab 14 \tab 40 \tab 28 \tab 206 \tab 27 \tab 12 \tab 23 \tab 52 \tab 71 \tab 6 \tab 36 \tab 224 \tab 68 \tab 178 \tab 168 \tab 21 \tab 44 \tab 21 \tab - \tab 2765 \cr
   2004 \tab 82 \tab 116 \tab 115 \tab 945 \tab 1 \tab 236 \tab 13 \tab 55 \tab 25 \tab 282 \tab 14 \tab 14 \tab 23 \tab 48 \tab 70 \tab 12 \tab 40 \tab 179 \tab 86 \tab 181 \tab 166 \tab 36 \tab 39 \tab 34 \tab - \tab 2812 \cr
   2006 \tab 117 \tab 114 \tab 93 \tab 971 \tab 7 \tab 233 \tab 22 \tab 42 \tab 37 \tab 261 \tab 32 \tab 18 \tab 29 \tab 69 \tab 109 \tab 14 \tab 36 \tab 222 \tab 78 \tab 167 \tab 1687 \tab 31 \tab 94 \tab 27 \tab - \tab 4510 \cr
   2008 \tab 73 \tab 100 \tab 65 \tab 725 \tab 5 \tab 143 \tab 7 \tab 27 \tab 19 \tab 182 \tab 17 \tab 12 \tab 23 \tab 50 \tab 69 \tab 1 \tab 25 \tab 128 \tab 43 \tab 156 \tab 83 \tab 4 \tab 52 \tab 14 \tab - \tab 2023 \cr
   2010 \tab 72 \tab 87 \tab 68 \tab 716 \tab 4 \tab 144 \tab 9 \tab 29 \tab 21 \tab 191 \tab 24 \tab 20 \tab 20 \tab 46 \tab 58 \tab 8 \tab 24 \tab 153 \tab 47 \tab 160 \tab 78 \tab 5 \tab 41 \tab 19 \tab - \tab 2044 \cr
   2012 \tab 68 \tab 68 \tab 71 \tab 680 \tab 3 \tab 158 \tab 6 \tab 36 \tab 26 \tab 185 \tab 15 \tab 12 \tab 21 \tab 51 \tab 73 \tab 8 \tab 24 \tab 113 \tab 50 \tab 148 \tab 74 \tab 7 \tab 49 \tab 28 \tab - \tab 1974 \cr
   2014 \tab 76 \tab 101 \tab 84 \tab 902 \tab 3 \tab 223 \tab 10 \tab 38 \tab 32 \tab 230 \tab 31 \tab 18 \tab 29 \tab 77 \tab 64 \tab 15 \tab 27 \tab 146 \tab 63 \tab 181 \tab 83 \tab 3 \tab 68 \tab 34 \tab - \tab 2538 \cr
   2016 \tab 100 \tab 98 \tab 86 \tab 1055 \tab 5 \tab 236 \tab 11 \tab 53 \tab 27 \tab 302 \tab 16 \tab 20 \tab 32 \tab 89 \tab 77 \tab 10 \tab 29 \tab 182 \tab 61 \tab 180 \tab 101 \tab 5 \tab 59 \tab 33 \tab - \tab 2867 \cr
   2018 \tab 71 \tab 70 \tab 69 \tab 826 \tab 6 \tab 208 \tab 10 \tab 43 \tab 26 \tab 278 \tab 19 \tab 24 \tab 33 \tab 59 \tab 67 \tab 7 \tab 28 \tab 98 \tab 65 \tab 169 \tab 84 \tab 6 \tab 52 \tab 30 \tab - \tab 2348 \cr
   2021 \tab 163 \tab 146 \tab 79 \tab 1580 \tab 4 \tab 374 \tab 7 \tab 84 \tab 15 \tab 517 \tab 17 \tab 48 \tab 30 \tab 106 \tab 57 \tab 15 \tab 34 \tab 144 \tab 71 \tab 316 \tab 100 \tab 3 \tab 55 \tab 67 \tab - \tab 4032 \cr
   2022 \tab 120 \tab 118 \tab 87 \tab 1339 \tab 2 \tab 272 \tab 7 \tab 55 \tab 31 \tab 470 \tab 23 \tab 40 \tab 13 \tab 109 \tab 81 \tab 11 \tab 26 \tab 120 \tab 68 \tab 326 \tab 98 \tab - \tab 50 \tab 78 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 2118 \tab 3137 \tab 2331 \tab 22815 \tab 100 \tab 4566 \tab 322 \tab 957 \tab 890 \tab 5226 \tab 928 \tab 406 \tab 1002 \tab 1146 \tab 2312 \tab 165 \tab 1267 \tab 6883 \tab 1928 \tab 7077 \tab 4246 \tab 464 \tab 1578 \tab 526 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{age}},
\code{\link{agekdbrn}},
\code{\link{codeg}},
\code{\link{coeduc}},
\code{\link{degree}},
\code{\link{educ}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{speduc}}

Other Education: 
\code{\link{degree}},
\code{\link{educ}},
\code{\link{padeg}},
\code{\link{paeduc}},
\code{\link{speduc}}
}
\concept{Core}
\concept{Education}
\keyword{variable}
