% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{mafund}
\alias{mafund}
\title{How fundamentalist was r's mother}
\source{
General Social Survey https://gss.norc.org
}
\description{
mafund
}
\details{
Question C. Fundamentalism/Liberalism of Mother's Religion
}
\section{Values}{

\itemize{
\item \code{1} fundamentalist
\item \code{2} moderate
\item \code{3} liberal
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1062/vshow}{GSS Data Explorer page for \code{mafund}}.
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}

Other Religion: 
\code{\link{churhsch}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
