% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{maisco88}
\alias{maisco88}
\title{R's mothers occupation, 1980 census & 1988 isco code}
\source{
General Social Survey https://gss.norc.org
}
\description{
maisco88
}
\details{
Question Respondent's Mother's Occupation, 1988 Census
}
\section{Values}{

\itemize{
\item \code{9997} uncodeable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{isco08}},
\code{\link{isco88}},
\code{\link{maisc681}},
\code{\link{maisco08}},
\code{\link{paisco08}},
\code{\link{paisco88}},
\code{\link{spisc681}},
\code{\link{spisco08}},
\code{\link{spisco88}}
}
\concept{Occupation}
\keyword{variable}
