% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{major1}
\alias{major1}
\title{College major 1}
\source{
General Social Survey https://gss.norc.org
}
\description{
major1
}
\details{
Question What was your major or field of study when you received your (RESPONDENT'S COLLEGE DEGREE) degree? IF THE RESPONDENT RECEIVED MORE THAN ONE GRADUATE LEVEL DEGREE, ASK ABOUT THE HIGHEST DEGREE OBTAINED (E.G. ABOUT A Ph.D. RATHER THAN AN M.B.A).
(COUNT UP TO 2 MENTIONS)
}
\section{Values}{

\itemize{
\item \verb{[1]} accounting/bookkeeping
\item \verb{[2]} advertising
\item \verb{[3]} agriculture/horticulture
\item \verb{[4]} allied health
\item \verb{[5]} anthropology
\item \verb{[6]} architecture
\item \verb{[7]} art
\item \verb{[8]} biology
\item \verb{[9]} business administration
\item \verb{[11]} chemistry
\item \verb{[12]} communications/speech
\item \verb{[13]} comm. disorders
\item \verb{[14]} computer science
\item \verb{[15]} dentistry
\item \verb{[16]} education
\item \verb{[17]} economics
\item \verb{[18]} engineering
\item \verb{[19]} english
\item \verb{[20]} finance
\item \verb{[21]} foreign language
\item \verb{[22]} forestry
\item \verb{[23]} geography
\item \verb{[24]} geology
\item \verb{[25]} history
\item \verb{[26]} home economics
\item \verb{[27]} industry & technology
\item \verb{[28]} journalism
\item \verb{[29]} law
\item \verb{[30]} law enforcement
\item \verb{[31]} library science
\item \verb{[32]} marketing
\item \verb{[33]} mathematics
\item \verb{[34]} medicine
\item \verb{[35]} music
\item \verb{[36]} nursing
\item \verb{[37]} optometry
\item \verb{[38]} pharmacy
\item \verb{[39]} philosophy
\item \verb{[40]} physical education
\item \verb{[41]} physics
\item \verb{[42]} psychology
\item \verb{[43]} political science/international relations
\item \verb{[44]} sociology
\item \verb{[45]} special education
\item \verb{[46]} theater arts
\item \verb{[47]} theology
\item \verb{[48]} veterinary medicine
\item \verb{[49]} liberal arts
\item \verb{[50]} other
\item \verb{[51]} general sciences
\item \verb{[52]} social work
\item \verb{[53]} general studies
\item \verb{[54]} other vocational
\item \verb{[55]} health
\item \verb{[56]} industrial relations
\item \verb{[57]} child/human/family development
\item \verb{[58]} food science/nutrition/culinary arts
\item \verb{[59]} environmental science/ecology
\item \verb{[60]} social sciences
\item \verb{[61]} human services/human resources
\item \verb{[62]} visual arts/graphic design/design and drafting
\item \verb{[63]} fine arts
\item \verb{[64]} humanities
\item \verb{[65]} ethnic studies
\item \verb{[66]} educational administration
\item \verb{[67]} television/film
\item \verb{[68]} aviation/aeronatics
\item \verb{[69]} statistics/biostatistics
\item \verb{[70]} criminology/criminal justice
\item \verb{[71]} administrative science/public administration
\item \verb{[72]} electronics
\item \verb{[73]} urban and regional planning
\item \verb{[74]} mechanics/machine trade
\item \verb{[75]} dance
\item \verb{[76]} gerontology
\item \verb{[77]} public relations
\item \verb{[78]} textiles/cloth
\item \verb{[79]} parks and recreation
\item \verb{[80]} information technology
\item \verb{[81]} fashion
\item \verb{[82]} counseling
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/63/vshow}{GSS Data Explorer page for \code{major1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab accounting/bookkeeping \tab advertising \tab agriculture/horticulture \tab anthropology \tab architecture \tab art \tab biology \tab business administration \tab chemistry \tab communications/speech \tab comm. disorders \tab computer science \tab dentistry \tab education \tab economics \tab engineering \tab english \tab finance \tab foreign language \tab forestry \tab geology \tab history \tab industry & technology \tab journalism \tab law \tab law enforcement \tab library science \tab marketing \tab mathematics \tab medicine \tab music \tab nursing \tab pharmacy \tab philosophy \tab physical education \tab physics \tab psychology \tab political science/international relations \tab sociology \tab special education \tab theater arts \tab theology \tab veterinary medicine \tab liberal arts \tab other \tab general sciences \tab social work \tab general studies \tab other vocational \tab health \tab industrial relations \tab child/human/family development \tab food science/nutrition/culinary arts \tab environmental science/ecology \tab social sciences \tab human services/human resources \tab visual arts/graphic design/design and drafting \tab fine arts \tab ethnic studies \tab educational administration \tab aviation/aeronatics \tab criminology/criminal justice \tab administrative science/public administration \tab electronics \tab urban and regional planning \tab dance \tab gerontology \tab public relations \tab textiles/cloth \tab parks and recreation \tab don't know \tab allied health \tab geography \tab humanities \tab television/film \tab mechanics/machine trade \tab information technology \tab no answer \tab home economics \tab statistics/biostatistics \tab fashion \tab counseling \tab optometry \tab skipped on web \tab not available in this year \tab Total \cr
   2012 \tab 1263 \tab 28 \tab 3 \tab 8 \tab 3 \tab 2 \tab 6 \tab 16 \tab 90 \tab 5 \tab 20 \tab 4 \tab 25 \tab 2 \tab 73 \tab 11 \tab 47 \tab 23 \tab 7 \tab 4 \tab 1 \tab 1 \tab 10 \tab 3 \tab 5 \tab 13 \tab 3 \tab 4 \tab 11 \tab 5 \tab 9 \tab 4 \tab 36 \tab 2 \tab 2 \tab 9 \tab 3 \tab 32 \tab 16 \tab 9 \tab 5 \tab 6 \tab 6 \tab 1 \tab 8 \tab 8 \tab 10 \tab 7 \tab 2 \tab 5 \tab 23 \tab 1 \tab 11 \tab 3 \tab 5 \tab 4 \tab 3 \tab 3 \tab 4 \tab 3 \tab 3 \tab 2 \tab 13 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1597 \tab 32 \tab 2 \tab 2 \tab 5 \tab 3 \tab 7 \tab 22 \tab 142 \tab 8 \tab 18 \tab 6 \tab 24 \tab 4 \tab 91 \tab 19 \tab 49 \tab 26 \tab 15 \tab 8 \tab - \tab 3 \tab 19 \tab 4 \tab 6 \tab 18 \tab 5 \tab 5 \tab 15 \tab 10 \tab 25 \tab 2 \tab 39 \tab 5 \tab - \tab 6 \tab 6 \tab 32 \tab 22 \tab 15 \tab 3 \tab 2 \tab 6 \tab 5 \tab 16 \tab 10 \tab 13 \tab 17 \tab 5 \tab 11 \tab 31 \tab - \tab 3 \tab 6 \tab 5 \tab 2 \tab 7 \tab 8 \tab 5 \tab 1 \tab 4 \tab 1 \tab 17 \tab 11 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1795 \tab 30 \tab - \tab 7 \tab 1 \tab 5 \tab 11 \tab 33 \tab 172 \tab 10 \tab 26 \tab 2 \tab 33 \tab 3 \tab 97 \tab 13 \tab 47 \tab 27 \tab 14 \tab 6 \tab 3 \tab 4 \tab 14 \tab 6 \tab 6 \tab 23 \tab 4 \tab 2 \tab 13 \tab 12 \tab 12 \tab 10 \tab 60 \tab 4 \tab 2 \tab 16 \tab 7 \tab 34 \tab 19 \tab 10 \tab 5 \tab 3 \tab 13 \tab 3 \tab 12 \tab 21 \tab 15 \tab 24 \tab 7 \tab 6 \tab 31 \tab - \tab 7 \tab 9 \tab 6 \tab 7 \tab 7 \tab 9 \tab 5 \tab - \tab 8 \tab 1 \tab 17 \tab 3 \tab 5 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 6 \tab 1 \tab 8 \tab 5 \tab 3 \tab 2 \tab 3 \tab 11 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1435 \tab 29 \tab - \tab 5 \tab 1 \tab 3 \tab 10 \tab 26 \tab 138 \tab 4 \tab 18 \tab 2 \tab 17 \tab 5 \tab 79 \tab 19 \tab 54 \tab 24 \tab 16 \tab 5 \tab - \tab 2 \tab 19 \tab - \tab 4 \tab 14 \tab 2 \tab 3 \tab 12 \tab 5 \tab 11 \tab 2 \tab 51 \tab 4 \tab 2 \tab 6 \tab 4 \tab 29 \tab 14 \tab 12 \tab 2 \tab 1 \tab 8 \tab 4 \tab 10 \tab 27 \tab 14 \tab 7 \tab 7 \tab 5 \tab 42 \tab 3 \tab 7 \tab 9 \tab 8 \tab 5 \tab 5 \tab 10 \tab 6 \tab - \tab 9 \tab 3 \tab 13 \tab 5 \tab 9 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 4 \tab 11 \tab 3 \tab 2 \tab 2 \tab 1 \tab 9 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 70 \tab 1 \tab 12 \tab 6 \tab 16 \tab 24 \tab 64 \tab 309 \tab 17 \tab 47 \tab - \tab 67 \tab 6 \tab 186 \tab 23 \tab 103 \tab 44 \tab 20 \tab 19 \tab 1 \tab 6 \tab 31 \tab 1 \tab 16 \tab 45 \tab 4 \tab 11 \tab 20 \tab 17 \tab 85 \tab 13 \tab 93 \tab 13 \tab 6 \tab 3 \tab 5 \tab 63 \tab 41 \tab 26 \tab 11 \tab 10 \tab 23 \tab 6 \tab 21 \tab 40 \tab 27 \tab 28 \tab 18 \tab 3 \tab 31 \tab - \tab 13 \tab 5 \tab 21 \tab 14 \tab 16 \tab 3 \tab 35 \tab - \tab 26 \tab 3 \tab 58 \tab 25 \tab 23 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 28 \tab 12 \tab 11 \tab 3 \tab 5 \tab 10 \tab 28 \tab 12 \tab 4 \tab 4 \tab 6 \tab 16 \tab 4 \tab 30 \tab 1856 \tab 4032 \cr
   2022 \tab 2013 \tab 41 \tab 2 \tab 7 \tab 2 \tab 8 \tab 21 \tab 47 \tab 221 \tab 13 \tab 32 \tab 4 \tab 48 \tab 5 \tab 129 \tab 13 \tab 80 \tab 41 \tab 22 \tab 6 \tab - \tab 2 \tab 17 \tab 6 \tab 12 \tab 32 \tab 1 \tab 4 \tab 12 \tab 21 \tab 23 \tab 7 \tab 90 \tab 6 \tab 4 \tab 9 \tab 7 \tab 60 \tab 34 \tab 19 \tab 8 \tab 4 \tab 13 \tab 5 \tab 17 \tab 18 \tab 27 \tab 2 \tab 14 \tab 6 \tab 50 \tab 1 \tab 14 \tab 13 \tab 14 \tab 10 \tab 17 \tab 12 \tab 9 \tab 1 \tab 8 \tab 6 \tab 32 \tab 7 \tab 13 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 5 \tab 8 \tab 10 \tab 14 \tab 3 \tab - \tab 2 \tab 27 \tab 1 \tab 20 \tab - \tab 3544 \cr
   Total \tab 8103 \tab 230 \tab 8 \tab 41 \tab 18 \tab 37 \tab 79 \tab 208 \tab 1072 \tab 57 \tab 161 \tab 18 \tab 214 \tab 25 \tab 655 \tab 98 \tab 380 \tab 185 \tab 94 \tab 48 \tab 5 \tab 18 \tab 110 \tab 20 \tab 49 \tab 145 \tab 19 \tab 29 \tab 83 \tab 70 \tab 165 \tab 38 \tab 369 \tab 34 \tab 16 \tab 49 \tab 32 \tab 250 \tab 146 \tab 91 \tab 34 \tab 26 \tab 69 \tab 24 \tab 84 \tab 124 \tab 106 \tab 85 \tab 53 \tab 36 \tab 208 \tab 5 \tab 55 \tab 45 \tab 59 \tab 42 \tab 55 \tab 45 \tab 64 \tab 5 \tab 58 \tab 16 \tab 150 \tab 53 \tab 62 \tab 13 \tab 4 \tab 5 \tab 10 \tab 9 \tab 6 \tab 34 \tab 15 \tab 22 \tab 7 \tab 19 \tab 24 \tab 62 \tab 35 \tab 12 \tab 8 \tab 12 \tab 63 \tab 5 \tab 50 \tab 1856 \tab 17303 \cr
}
}

\seealso{
Other Core: 
\code{\link{codipged}},
\code{\link{cosector}},
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major2}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\keyword{variable}
