% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_11.R
\name{major2}
\alias{major2}
\title{College major 2}
\source{
General Social Survey https://gss.norc.org
}
\description{
major2
}
\details{
Question What was your major or field of study when you received your (RESPONDENT'S COLLEGE DEGREE) degree? IF THE RESPONDENT RECEIVED MORE THAN ONE GRADUATE LEVEL DEGREE, ASK ABOUT THE HIGHEST DEGREE OBTAINED (E.G. ABOUT A Ph.D. RATHER THAN AN M.B.A).
(COUNT UP TO 2 MENTIONS)
}
\section{Values}{

\itemize{
\item \verb{[1]} accounting/bookkeeping
\item \verb{[2]} advertising
\item \verb{[3]} agriculture/horticulture
\item \verb{[4]} allied health
\item \verb{[5]} anthropology
\item \verb{[6]} architecture
\item \verb{[7]} art
\item \verb{[8]} biology
\item \verb{[9]} business administration
\item \verb{[11]} chemistry
\item \verb{[12]} communications/speech
\item \verb{[13]} comm. disorders
\item \verb{[14]} computer science
\item \verb{[15]} dentistry
\item \verb{[16]} education
\item \verb{[17]} economics
\item \verb{[18]} engineering
\item \verb{[19]} english
\item \verb{[20]} finance
\item \verb{[21]} foreign language
\item \verb{[22]} forestry
\item \verb{[23]} geography
\item \verb{[24]} geology
\item \verb{[25]} history
\item \verb{[26]} home economics
\item \verb{[27]} industry & technology
\item \verb{[28]} journalism
\item \verb{[29]} law
\item \verb{[30]} law enforcement
\item \verb{[31]} library science
\item \verb{[32]} marketing
\item \verb{[33]} mathematics
\item \verb{[34]} medicine
\item \verb{[35]} music
\item \verb{[36]} nursing
\item \verb{[37]} optometry
\item \verb{[38]} pharmacy
\item \verb{[39]} philosophy
\item \verb{[40]} physical education
\item \verb{[41]} physics
\item \verb{[42]} psychology
\item \verb{[43]} political science/international relations
\item \verb{[44]} sociology
\item \verb{[45]} special education
\item \verb{[46]} theater arts
\item \verb{[47]} theology
\item \verb{[48]} veterinary medicine
\item \verb{[49]} liberal arts
\item \verb{[50]} other
\item \verb{[51]} general sciences
\item \verb{[52]} social work
\item \verb{[53]} general studies
\item \verb{[54]} other vocational
\item \verb{[55]} health
\item \verb{[56]} industrial relations
\item \verb{[57]} child/human/family development
\item \verb{[58]} food science/nutrition/culinary arts
\item \verb{[59]} environmental science/ecology
\item \verb{[60]} social sciences
\item \verb{[61]} human services/human resources
\item \verb{[62]} visual arts/graphic design/design and drafting
\item \verb{[63]} fine arts
\item \verb{[64]} humanities
\item \verb{[65]} ethnic studies
\item \verb{[66]} educational administration
\item \verb{[67]} television/film
\item \verb{[68]} aviation/aeronatics
\item \verb{[69]} statistics/biostatistics
\item \verb{[70]} criminology/criminal justice
\item \verb{[71]} administrative science/public administration
\item \verb{[72]} electronics
\item \verb{[73]} urban and regional planning
\item \verb{[74]} mechanics/machine trade
\item \verb{[75]} dance
\item \verb{[76]} gerontology
\item \verb{[77]} public relations
\item \verb{[78]} textiles/cloth
\item \verb{[79]} parks and recreation
\item \verb{[80]} information technology
\item \verb{[81]} fashion
\item \verb{[82]} counseling
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/64/vshow}{GSS Data Explorer page for \code{major2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab accounting/bookkeeping \tab art \tab biology \tab business administration \tab chemistry \tab computer science \tab dentistry \tab education \tab economics \tab english \tab finance \tab foreign language \tab geology \tab history \tab industry & technology \tab law \tab marketing \tab mathematics \tab medicine \tab music \tab nursing \tab pharmacy \tab philosophy \tab psychology \tab political science/international relations \tab sociology \tab special education \tab theater arts \tab theology \tab other \tab social work \tab child/human/family development \tab environmental science/ecology \tab social sciences \tab human services/human resources \tab visual arts/graphic design/design and drafting \tab television/film \tab aviation/aeronatics \tab statistics/biostatistics \tab urban and regional planning \tab textiles/cloth \tab parks and recreation \tab don't know \tab no answer \tab anthropology \tab communications/speech \tab comm. disorders \tab engineering \tab forestry \tab journalism \tab law enforcement \tab library science \tab physical education \tab physics \tab general studies \tab other vocational \tab health \tab fine arts \tab educational administration \tab criminology/criminal justice \tab information technology \tab agriculture/horticulture \tab geography \tab optometry \tab liberal arts \tab general sciences \tab food science/nutrition/culinary arts \tab humanities \tab ethnic studies \tab administrative science/public administration \tab mechanics/machine trade \tab dance \tab public relations \tab counseling \tab gerontology \tab not available in this year \tab advertising \tab architecture \tab home economics \tab skipped on web \tab Total \cr
   2012 \tab 1852 \tab 2 \tab 2 \tab 2 \tab 15 \tab 3 \tab 5 \tab 1 \tab 9 \tab 6 \tab 6 \tab 4 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 6 \tab 4 \tab 4 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2345 \tab 6 \tab 6 \tab 2 \tab 27 \tab 2 \tab 3 \tab - \tab 16 \tab 3 \tab 4 \tab 5 \tab 5 \tab 1 \tab 6 \tab - \tab 3 \tab 4 \tab 2 \tab 2 \tab 5 \tab 2 \tab - \tab 1 \tab 10 \tab 5 \tab 2 \tab 2 \tab 2 \tab 2 \tab 5 \tab 2 \tab 4 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 13 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2643 \tab 6 \tab 5 \tab 7 \tab 29 \tab 3 \tab 7 \tab 1 \tab 19 \tab 10 \tab 5 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 5 \tab 8 \tab 5 \tab 1 \tab 4 \tab - \tab - \tab 9 \tab 3 \tab 5 \tab 4 \tab 3 \tab 3 \tab 4 \tab 3 \tab 3 \tab 2 \tab 2 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 5 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 11 \tab - \tab 1 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2201 \tab 9 \tab 1 \tab 4 \tab 9 \tab 4 \tab 10 \tab 1 \tab 15 \tab 4 \tab 7 \tab 4 \tab 3 \tab - \tab 8 \tab - \tab 2 \tab 6 \tab 11 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 2941 \tab 20 \tab 15 \tab 14 \tab 58 \tab 9 \tab 8 \tab 1 \tab 37 \tab 16 \tab 25 \tab 14 \tab 15 \tab 2 \tab 14 \tab 6 \tab 6 \tab 10 \tab 9 \tab 5 \tab 4 \tab 9 \tab 2 \tab 2 \tab 20 \tab 11 \tab 7 \tab 8 \tab 1 \tab 8 \tab 9 \tab 3 \tab 9 \tab 2 \tab 2 \tab 13 \tab 11 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab 15 \tab 30 \tab 5 \tab 14 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 1 \tab 8 \tab 1 \tab 1 \tab 3 \tab 20 \tab - \tab - \tab 13 \tab 4 \tab 3 \tab - \tab - \tab 6 \tab 8 \tab 3 \tab - \tab 5 \tab 6 \tab - \tab - \tab 2 \tab 8 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 20 \tab 3544 \cr
   Total \tab 11982 \tab 43 \tab 29 \tab 29 \tab 138 \tab 21 \tab 33 \tab 4 \tab 96 \tab 39 \tab 47 \tab 29 \tab 27 \tab 4 \tab 35 \tab 7 \tab 14 \tab 29 \tab 33 \tab 14 \tab 13 \tab 21 \tab 3 \tab 4 \tab 45 \tab 24 \tab 19 \tab 15 \tab 8 \tab 16 \tab 30 \tab 9 \tab 18 \tab 9 \tab 7 \tab 24 \tab 15 \tab 4 \tab 3 \tab 6 \tab 3 \tab 1 \tab 1 \tab 18 \tab 32 \tab 8 \tab 23 \tab 2 \tab 16 \tab 2 \tab 7 \tab 1 \tab 6 \tab 12 \tab 3 \tab 2 \tab 7 \tab 49 \tab 4 \tab 3 \tab 23 \tab 9 \tab 5 \tab 3 \tab 1 \tab 10 \tab 12 \tab 4 \tab 2 \tab 7 \tab 7 \tab 2 \tab 1 \tab 3 \tab 10 \tab 2 \tab 0 \tab 2 \tab 1 \tab 1 \tab 20 \tab 13271 \cr
}
}

\seealso{
Other Core: 
\code{\link{codipged}},
\code{\link{cosector}},
\code{\link{dipged}},
\code{\link{madeg}},
\code{\link{major1}},
\code{\link{spdeg}},
\code{\link{spdipged}},
\code{\link{whencol}},
\code{\link{whenhs}}
}
\concept{Core}
\keyword{variable}
