% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{majorcol}
\alias{majorcol}
\title{The field of degree r earned}
\source{
General Social Survey https://gss.norc.org
}
\description{
majorcol
}
\details{
Question In what field was that degree?
}
\section{Values}{

\itemize{
\item \code{1} accounting/bookkeeping
\item \code{2} advertising
\item \code{3} agriculture/horticulture
\item \code{4} allied health
\item \code{5} anthropology
\item \code{6} architecture
\item \code{7} art
\item \code{8} biology
\item \code{9} business administration
\item \code{11} chemistry
\item \code{12} communications/speech
\item \code{13} comm. disorders
\item \code{14} computer science
\item \code{15} dentistry
\item \code{16} education
\item \code{17} economics
\item \code{18} engineering
\item \code{19} english
\item \code{20} finance
\item \code{21} foreign language
\item \code{22} forestry
\item \code{23} geography
\item \code{24} geology
\item \code{25} history
\item \code{26} home economics
\item \code{27} industry & technology
\item \code{28} journalism
\item \code{29} law
\item \code{30} law enforcement
\item \code{31} library science
\item \code{32} marketing
\item \code{33} mathematics
\item \code{34} medicine
\item \code{35} music
\item \code{36} nursing
\item \code{37} optometry
\item \code{38} pharmacy
\item \code{39} philosophy
\item \code{40} physical education
\item \code{41} physics
\item \code{42} psychology
\item \code{43} political science/international relations
\item \code{44} sociology
\item \code{45} special education
\item \code{46} theater arts
\item \code{47} theology
\item \code{48} veterinary medicine
\item \code{49} liberal arts
\item \code{50} other
\item \code{51} general sciences
\item \code{52} social work
\item \code{53} general studies
\item \code{54} other vocational
\item \code{55} health
\item \code{56} industrial relations
\item \code{57} child/human/family development
\item \code{58} food science/nutrition/culinary arts
\item \code{59} environmental science/ecology
\item \code{60} social sciences
\item \code{61} human services/human resources
\item \code{62} visual arts/graphic design/design and drafting
\item \code{63} fine arts
\item \code{64} humanities
\item \code{65} ethnic studies
\item \code{66} educational administration
\item \code{67} television/film
\item \code{68} aviation/aeronatics
\item \code{69} statistics/biostatistics
\item \code{70} criminology/criminal justice
\item \code{71} administrative science/public administration
\item \code{72} electronics
\item \code{73} urban and regional planning
\item \code{74} mechanics/machine trade
\item \code{75} dance
\item \code{76} gerontology
\item \code{77} public relations
\item \code{78} textiles/cloth
\item \code{79} parks and recreation
\item \code{80} information technology
\item \code{81} fashion
\item \code{82} counseling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3348/vshow}{GSS Data Explorer page for \code{majorcol}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab accounting/bookkeeping \tab agriculture/horticulture \tab anthropology \tab architecture \tab art \tab biology \tab business administration \tab chemistry \tab communications/speech \tab comm. disorders \tab computer science \tab dentistry \tab education \tab economics \tab engineering \tab english \tab finance \tab foreign language \tab forestry \tab geography \tab geology \tab history \tab industry & technology \tab journalism \tab law \tab law enforcement \tab library science \tab marketing \tab mathematics \tab medicine \tab music \tab nursing \tab pharmacy \tab philosophy \tab physical education \tab physics \tab psychology \tab political science/international relations \tab sociology \tab special education \tab theater arts \tab theology \tab veterinary medicine \tab liberal arts \tab other \tab general sciences \tab social work \tab general studies \tab other vocational \tab health \tab don't know \tab no answer \tab allied health \tab home economics \tab industrial relations \tab child/human/family development \tab food science/nutrition/culinary arts \tab environmental science/ecology \tab human services/human resources \tab humanities \tab ethnic studies \tab television/film \tab statistics/biostatistics \tab advertising \tab visual arts/graphic design/design and drafting \tab fine arts \tab criminology/criminal justice \tab administrative science/public administration \tab electronics \tab dance \tab gerontology \tab textiles/cloth \tab parks and recreation \tab social sciences \tab educational administration \tab urban and regional planning \tab public relations \tab information technology \tab fashion \tab counseling \tab aviation/aeronatics \tab mechanics/machine trade \tab Total \cr
   2006 \tab 3375 \tab 28 \tab 3 \tab 4 \tab 3 \tab 15 \tab 23 \tab 117 \tab 4 \tab 25 \tab 2 \tab 28 \tab 2 \tab 87 \tab 11 \tab 57 \tab 17 \tab 14 \tab 9 \tab 1 \tab 3 \tab 3 \tab 7 \tab 5 \tab 3 \tab 29 \tab 4 \tab 4 \tab 14 \tab 11 \tab 8 \tab 4 \tab 51 \tab 2 \tab 2 \tab 5 \tab 5 \tab 34 \tab 14 \tab 7 \tab 2 \tab 2 \tab 6 \tab 1 \tab 14 \tab 26 \tab 21 \tab 10 \tab 24 \tab 65 \tab 25 \tab 47 \tab 227 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1493 \tab 13 \tab 2 \tab 3 \tab 5 \tab 14 \tab 16 \tab 79 \tab 1 \tab 11 \tab 1 \tab 22 \tab 1 \tab 52 \tab 6 \tab 27 \tab 10 \tab 6 \tab 3 \tab 1 \tab 2 \tab 4 \tab 8 \tab 2 \tab 3 \tab 16 \tab 11 \tab 3 \tab 9 \tab 5 \tab 5 \tab 8 \tab 24 \tab - \tab 2 \tab 2 \tab 2 \tab 16 \tab 14 \tab 7 \tab - \tab 1 \tab 3 \tab 3 \tab 10 \tab 10 \tab 16 \tab 8 \tab 10 \tab 30 \tab 15 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1772 \tab 4 \tab 1 \tab - \tab 3 \tab 5 \tab 10 \tab 34 \tab 3 \tab 9 \tab - \tab 3 \tab 1 \tab 35 \tab 3 \tab 15 \tab 10 \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 6 \tab 4 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 18 \tab - \tab - \tab - \tab 2 \tab 10 \tab 5 \tab 5 \tab 1 \tab 3 \tab 2 \tab 2 \tab 8 \tab 10 \tab 6 \tab 3 \tab 1 \tab 4 \tab 7 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1612 \tab 10 \tab 1 \tab 2 \tab 2 \tab 3 \tab 6 \tab 54 \tab 1 \tab 10 \tab - \tab 14 \tab 2 \tab 46 \tab 5 \tab 22 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab 3 \tab 4 \tab 9 \tab - \tab 1 \tab 5 \tab 4 \tab 5 \tab 4 \tab 22 \tab 2 \tab - \tab 5 \tab 1 \tab 9 \tab 10 \tab 6 \tab 3 \tab 2 \tab 2 \tab - \tab 4 \tab 3 \tab 5 \tab 5 \tab - \tab 3 \tab 11 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 9 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2099 \tab 14 \tab - \tab 2 \tab 2 \tab 3 \tab 12 \tab 64 \tab 2 \tab 10 \tab 1 \tab 12 \tab - \tab 50 \tab 5 \tab 21 \tab 9 \tab 6 \tab 1 \tab - \tab 1 \tab 2 \tab 9 \tab 3 \tab 2 \tab 11 \tab 4 \tab 2 \tab 11 \tab 4 \tab 11 \tab 1 \tab 17 \tab - \tab - \tab 3 \tab 2 \tab 10 \tab 10 \tab 6 \tab - \tab - \tab 3 \tab 1 \tab 7 \tab 6 \tab 8 \tab 9 \tab 7 \tab 6 \tab 18 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 5 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 3 \tab 9 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2338 \tab 14 \tab 2 \tab 1 \tab 2 \tab 2 \tab 11 \tab 101 \tab 5 \tab 14 \tab - \tab 19 \tab 2 \tab 52 \tab 7 \tab 21 \tab 16 \tab 7 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 4 \tab 14 \tab 3 \tab 2 \tab 3 \tab 6 \tab 8 \tab 4 \tab 28 \tab 1 \tab 1 \tab 4 \tab 3 \tab 19 \tab 12 \tab 6 \tab 3 \tab 2 \tab 3 \tab - \tab 6 \tab 7 \tab 9 \tab 8 \tab 3 \tab 3 \tab 13 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 4 \tab 4 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 4 \tab 6 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab 5 \tab 4 \tab 7 \tab - \tab - \tab 2867 \cr
   2018 \tab 1914 \tab 13 \tab 1 \tab 1 \tab 2 \tab 8 \tab 6 \tab 58 \tab 3 \tab 14 \tab 1 \tab 7 \tab 3 \tab 44 \tab 13 \tab 22 \tab 6 \tab 9 \tab 4 \tab - \tab 1 \tab 2 \tab 9 \tab - \tab 1 \tab 9 \tab - \tab - \tab 6 \tab 3 \tab 16 \tab 1 \tab 25 \tab 2 \tab 1 \tab 4 \tab 1 \tab 15 \tab 5 \tab 6 \tab - \tab 1 \tab 6 \tab 2 \tab 8 \tab 13 \tab 9 \tab 5 \tab - \tab 3 \tab 15 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 3 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 5 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 4 \tab 3 \tab 1 \tab 2348 \cr
   Total \tab 14603 \tab 96 \tab 10 \tab 13 \tab 19 \tab 50 \tab 84 \tab 507 \tab 19 \tab 93 \tab 5 \tab 105 \tab 11 \tab 366 \tab 50 \tab 185 \tab 74 \tab 52 \tab 23 \tab 5 \tab 9 \tab 12 \tab 46 \tab 15 \tab 18 \tab 94 \tab 26 \tab 15 \tab 49 \tab 34 \tab 58 \tab 23 \tab 185 \tab 7 \tab 6 \tab 23 \tab 16 \tab 113 \tab 70 \tab 43 \tab 9 \tab 11 \tab 25 \tab 9 \tab 57 \tab 75 \tab 74 \tab 48 \tab 45 \tab 114 \tab 104 \tab 55 \tab 239 \tab 4 \tab 5 \tab 2 \tab 13 \tab 14 \tab 11 \tab 18 \tab 3 \tab 3 \tab 7 \tab 2 \tab 2 \tab 11 \tab 12 \tab 29 \tab 6 \tab 11 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 5 \tab 4 \tab 2 \tab 12 \tab 4 \tab 11 \tab 3 \tab 1 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{solarrev}},
\code{\link{tomatoes}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
