% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{majorcol}
\alias{majorcol}
\title{The field of degree r earned}
\source{
General Social Survey https://gss.norc.org
}
\description{
majorcol
}
\details{
Question In what field was that degree?
}
\section{Values}{

\itemize{
\item \code{1} accounting/bookkeeping
\item \code{2} advertising
\item \code{3} agriculture/horticulture
\item \code{4} allied health
\item \code{5} anthropology
\item \code{6} architecture
\item \code{7} art
\item \code{8} biology
\item \code{9} business administration
\item \code{11} chemistry
\item \code{12} communications/speech
\item \code{13} comm. disorders
\item \code{14} computer science
\item \code{15} dentistry
\item \code{16} education
\item \code{17} economics
\item \code{18} engineering
\item \code{19} english
\item \code{20} finance
\item \code{21} foreign language
\item \code{22} forestry
\item \code{23} geography
\item \code{24} geology
\item \code{25} history
\item \code{26} home economics
\item \code{27} industry & technology
\item \code{28} journalism
\item \code{29} law
\item \code{30} law enforcement
\item \code{31} library science
\item \code{32} marketing
\item \code{33} mathematics
\item \code{34} medicine
\item \code{35} music
\item \code{36} nursing
\item \code{37} optometry
\item \code{38} pharmacy
\item \code{39} philosophy
\item \code{40} physical education
\item \code{41} physics
\item \code{42} psychology
\item \code{43} political science/international relations
\item \code{44} sociology
\item \code{45} special education
\item \code{46} theater arts
\item \code{47} theology
\item \code{48} veterinary medicine
\item \code{49} liberal arts
\item \code{50} other
\item \code{51} general sciences
\item \code{52} social work
\item \code{53} general studies
\item \code{54} other vocational
\item \code{55} health
\item \code{56} industrial relations
\item \code{57} child/human/family development
\item \code{58} food science/nutrition/culinary arts
\item \code{59} environmental science/ecology
\item \code{60} social sciences
\item \code{61} human services/human resources
\item \code{62} visual arts/graphic design/design and drafting
\item \code{63} fine arts
\item \code{64} humanities
\item \code{65} ethnic studies
\item \code{66} educational administration
\item \code{67} television/film
\item \code{68} aviation/aeronatics
\item \code{69} statistics/biostatistics
\item \code{70} criminology/criminal justice
\item \code{71} administrative science/public administration
\item \code{72} electronics
\item \code{73} urban and regional planning
\item \code{74} mechanics/machine trade
\item \code{75} dance
\item \code{76} gerontology
\item \code{77} public relations
\item \code{78} textiles/cloth
\item \code{79} parks and recreation
\item \code{80} information technology
\item \code{81} fashion
\item \code{82} counseling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3348/vshow}{GSS Data Explorer page for \code{majorcol}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{advfront}},
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{architct}},
\code{\link{artseals}},
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{bettrlfe}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{biosci}},
\code{\link{bizbstgm}},
\code{\link{bizbstgw}},
\code{\link{bizbstnk}},
\code{\link{bizbsttx}},
\code{\link{bizinfgm}},
\code{\link{bizinfgw}},
\code{\link{bizinfnk}},
\code{\link{bizinftx}},
\code{\link{biznuke}},
\code{\link{boyorgrl}},
\code{\link{caremost}},
\code{\link{chemgen}},
\code{\link{cloning}},
\code{\link{cmprgmng}},
\code{\link{coldeg1}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{comorsci}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{eatgm}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{econsci}},
\code{\link{electron}},
\code{\link{engagrnk}},
\code{\link{engbrng}},
\code{\link{engbstnk}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engnrsci}},
\code{\link{engnuke}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{enhinfnk}},
\code{\link{erosion}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{extinct}},
\code{\link{farming}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{gasregs}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{gmbiz}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{goldfish}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{h2olife}},
\code{\link{histsci}},
\code{\link{hotcore}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsmath}},
\code{\link{hsphys}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{indusgen}},
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intenvir}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intmil}},
\code{\link{intrhome}},
\code{\link{intsci}},
\code{\link{intspace}},
\code{\link{inttech}},
\code{\link{inuit}},
\code{\link{inuitway}},
\code{\link{journlsm}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lawenfrc}},
\code{\link{leadsci}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{marrcoun}},
\code{\link{medagrgm}},
\code{\link{medagrsc}},
\code{\link{medbstgm}},
\code{\link{medbstsc}},
\code{\link{medinfgm}},
\code{\link{medinfsc}},
\code{\link{medsci}},
\code{\link{medtreat}},
\code{\link{morempg}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{newsfrmy}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{noicecap}},
\code{\link{nosun}},
\code{\link{nukeelec}},
\code{\link{nukegen}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{penguins}},
\code{\link{physcsci}},
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}},
\code{\link{polbstgm}},
\code{\link{polbstgw}},
\code{\link{polbstnk}},
\code{\link{polbstsc}},
\code{\link{polbsttx}},
\code{\link{polinfgm}},
\code{\link{polinfgw}},
\code{\link{polinfnk}},
\code{\link{polinfsc}},
\code{\link{polinftx}},
\code{\link{polnuke}},
\code{\link{radioact}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{salth2o}},
\code{\link{sciagrgw}},
\code{\link{scibnfts}},
\code{\link{scibstgw}},
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientbr}},
\code{\link{scientda}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientmo}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientre}},
\code{\link{scientsn}},
\code{\link{scientwk}},
\code{\link{scifrom}},
\code{\link{scifromy}},
\code{\link{sciimp1}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{sciinfgo}},
\code{\link{sciinfgw}},
\code{\link{sciintro}},
\code{\link{scimath}},
\code{\link{scimode}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{scispec}},
\code{\link{scistudy}},
\code{\link{scitext}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}},
\code{\link{scresrch}},
\code{\link{sealevel}},
\code{\link{seeksci}},
\code{\link{seeksciy}},
\code{\link{seesand}},
\code{\link{slsmnshp}},
\code{\link{socsci}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen}},
\code{\link{tomatoes}},
\code{\link{toofast}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}},
\code{\link{txbiz}},
\code{\link{txeco}},
\code{\link{txpol}},
\code{\link{upbreath}},
\code{\link{viruses}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}},
\code{\link{watergen}},
\code{\link{weighing}},
\code{\link{whichsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
