% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{majorcol}
\alias{majorcol}
\title{The field of degree r earned}
\source{
General Social Survey https://gss.norc.org
}
\description{
majorcol
}
\details{
Question In what field was that degree?
}
\section{Values}{

\itemize{
\item \code{1} accounting/bookkeeping
\item \code{2} advertising
\item \code{3} agriculture/horticulture
\item \code{4} allied health
\item \code{5} anthropology
\item \code{6} architecture
\item \code{7} art
\item \code{8} biology
\item \code{9} business administration
\item \code{11} chemistry
\item \code{12} communications/speech
\item \code{13} comm. disorders
\item \code{14} computer science
\item \code{15} dentistry
\item \code{16} education
\item \code{17} economics
\item \code{18} engineering
\item \code{19} english
\item \code{20} finance
\item \code{21} foreign language
\item \code{22} forestry
\item \code{23} geography
\item \code{24} geology
\item \code{25} history
\item \code{26} home economics
\item \code{27} industry & technology
\item \code{28} journalism
\item \code{29} law
\item \code{30} law enforcement
\item \code{31} library science
\item \code{32} marketing
\item \code{33} mathematics
\item \code{34} medicine
\item \code{35} music
\item \code{36} nursing
\item \code{37} optometry
\item \code{38} pharmacy
\item \code{39} philosophy
\item \code{40} physical education
\item \code{41} physics
\item \code{42} psychology
\item \code{43} political science/international relations
\item \code{44} sociology
\item \code{45} special education
\item \code{46} theater arts
\item \code{47} theology
\item \code{48} veterinary medicine
\item \code{49} liberal arts
\item \code{50} other
\item \code{51} general sciences
\item \code{52} social work
\item \code{53} general studies
\item \code{54} other vocational
\item \code{55} health
\item \code{56} industrial relations
\item \code{57} child/human/family development
\item \code{58} food science/nutrition/culinary arts
\item \code{59} environmental science/ecology
\item \code{60} social sciences
\item \code{61} human services/human resources
\item \code{62} visual arts/graphic design/design and drafting
\item \code{63} fine arts
\item \code{64} humanities
\item \code{65} ethnic studies
\item \code{66} educational administration
\item \code{67} television/film
\item \code{68} aviation/aeronatics
\item \code{69} statistics/biostatistics
\item \code{70} criminology/criminal justice
\item \code{71} administrative science/public administration
\item \code{72} electronics
\item \code{73} urban and regional planning
\item \code{74} mechanics/machine trade
\item \code{75} dance
\item \code{76} gerontology
\item \code{77} public relations
\item \code{78} textiles/cloth
\item \code{79} parks and recreation
\item \code{80} information technology
\item \code{81} fashion
\item \code{82} counseling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3348/vshow}{GSS Data Explorer page for \code{majorcol}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{solarrev}},
\code{\link{tomatoes}}

Other Education: 
\code{\link{coldeg1}}
}
\concept{Education}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
