% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{majwoth}
\alias{majwoth}
\title{R's mom was jewish when r was growing up}
\source{
General Social Survey https://gss.norc.org
}
\description{
majwoth
}
\details{
Question Was you mother Jewish in any way when you were growing up?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} half/part
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3826/vshow}{GSS Data Explorer page for \code{majwoth}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab half/part \tab no \tab yes \tab Total \cr
   2008 \tab 2018 \tab 2 \tab 3 \tab - \tab 2023 \cr
   2012 \tab 1971 \tab - \tab 2 \tab 1 \tab 1974 \cr
   Total \tab 3989 \tab 2 \tab 5 \tab 1 \tab 3997 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{pajwoth}},
\code{\link{spjoth16}},
\code{\link{spjrel16}}
}
\concept{Jewish Identity}
\keyword{variable}
