% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{majwoth}
\alias{majwoth}
\title{R's mom was jewish when r was growing up}
\source{
General Social Survey https://gss.norc.org
}
\description{
majwoth
}
\details{
Question Was you mother Jewish in any way when you were growing up?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} half/part
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3826/vshow}{GSS Data Explorer page for \code{majwoth}}.
}

\seealso{
Other Jewish Identity: 
\code{\link{pajwoth}},
\code{\link{spjoth16}},
\code{\link{spjrel16}}

Other Jews: 
\code{\link{pajwoth}},
\code{\link{spjoth16}}

Other Religion: 
\code{\link{spjoth16}}
}
\concept{Jewish Identity}
\concept{Jews}
\concept{Religion}
\keyword{variable}
