% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_149.R
\name{makeart}
\alias{makeart}
\title{Made art or craft object in lst yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
makeart
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
C. Make art or craft objects such as pottery, woodworking, quilts, or paintings.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1407/vshow}{GSS Data Explorer page for \code{makeart}}.
}

\seealso{
Other Culture: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{conrock}},
\code{\link{hvymetal}},
\code{\link{oldies}},
\code{\link{opera}},
\code{\link{rap}},
\code{\link{reggae}},
\code{\link{visitart}}

Other Leisure: 
\code{\link{attsprts}},
\code{\link{autorace}},
\code{\link{visitart}}
}
\concept{Culture}
\concept{Leisure}
\keyword{variable}
