% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{makefrnd}
\alias{makefrnd}
\title{Religion helps people to make friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
makefrnd
}
\details{
Question Do you agree or disagree that…
B. ...that practicing a religion helps people to make friends?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4713/vshow}{GSS Data Explorer page for \code{makefrnd}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 321 \tab 799 \tab 160 \tab 59 \tab 11 \tab 10 \tab 5 \tab 2023 \cr
   2018 \tab 1173 \tab 228 \tab 625 \tab 191 \tab 96 \tab 17 \tab 6 \tab 12 \tab 2348 \cr
   Total \tab 1831 \tab 549 \tab 1424 \tab 351 \tab 155 \tab 28 \tab 16 \tab 17 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{rightpeo}},
\code{\link{vistholy}}
}
\concept{ISSP Religion}
\keyword{variable}
