% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_556.R
\name{maleornt}
\alias{maleornt}
\title{Interviewer coded male sexual orientation}
\source{
General Social Survey https://gss.norc.org
}
\description{
maleornt
}
\details{
Question INTERVIEWER ONLY, DO NOT ASK R
}
\section{Values}{

\itemize{
\item \verb{[1]} r is homosexual/gay
\item \verb{[2]} r is not homosexual/gay
\item \verb{[3]} r sexual orientation uncertain
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7294/vshow}{GSS Data Explorer page for \code{maleornt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab r is homosexual/gay \tab r is not homosexual/gay \tab r sexual orientation uncertain \tab no answer \tab Total \cr
   2018 \tab 1647 \tab 7 \tab 448 \tab 91 \tab 155 \tab 2348 \cr
   Total \tab 1647 \tab 7 \tab 448 \tab 91 \tab 155 \tab 2348 \cr
}
}

\keyword{variable}
