% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{manread}
\alias{manread}
\title{Should police open mail of a suspect?}
\source{
General Social Survey https://gss.norc.org
}
\description{
manread
}
\details{
Question Now, suppose the tip is about a man without a criminal record. Please circle a number for each question to show if you think police should be allowed without a court order to...
C. Open his mail.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4000/vshow}{GSS Data Explorer page for \code{manread}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 19 \tab 40 \tab 141 \tab 429 \tab 27 \tab 21 \tab 1534 \cr
   Total \tab 857 \tab 19 \tab 40 \tab 141 \tab 429 \tab 27 \tab 21 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{databank}},
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{progtax}},
\code{\link{verdict}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
