% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_274.R
\name{manual}
\alias{manual}
\title{Has r consulted manual for help with software}
\source{
General Social Survey https://gss.norc.org
}
\description{
manual
}
\details{
Question When you want to learn how to do something new with your software and you need advice on how to do it, which of the following do you do? (Name as many as are correct.) (PROMPT: Anything else?)
B. Look in a printed manual, book, or CD that tells how to use the software?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2493/vshow}{GSS Data Explorer page for \code{manual}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2075 \tab 486 \tab 168 \tab 88 \tab 2817 \cr
   Total \tab 2075 \tab 486 \tab 168 \tab 88 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{askother}},
\code{\link{askwork}},
\code{\link{callcomp}},
\code{\link{doonown}},
\code{\link{online}},
\code{\link{payother}},
\code{\link{wwwfrnd}},
\code{\link{wwwnei}},
\code{\link{wwwvol}}
}
\concept{Information Society}
\keyword{variable}
