% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{manvsemp}
\alias{manvsemp}
\title{Relations bw management and employees}
\source{
General Social Survey https://gss.norc.org
}
\description{
manvsemp
}
\details{
Question In general, how would you describe relations in your work place between management and employees?
}
\section{Values}{

\itemize{
\item \code{1} very good
\item \code{2} quite good
\item \code{3} neither good nor bad
\item \code{4} quite bad
\item \code{5} very bad
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2811/vshow}{GSS Data Explorer page for \code{manvsemp}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}

Other Job: 
\code{\link{cowrkhlp}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Job}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
