% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{mapres80}
\alias{mapres80}
\title{Mothers occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mapres80
}
\details{
Question A. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up?  That is, what was her job called?
OCCUPATION: (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What did she actually do in that job? Tell me, what were some of her main duties?
C. What kind of place did she work for?
INDUSTRY: (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was she self-employed, or did she work for someone else?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/46/vshow}{GSS Data Explorer page for \code{mapres80}}.
}

\seealso{
Other Core: 
\code{\link{maind80}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mawrkslf}},
\code{\link{paind10}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{papres10}},
\code{\link{papres105plus}}
}
\concept{Core}
\keyword{variable}
