% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{marelig}
\alias{marelig}
\title{What was mother's religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
marelig
}
\details{
Question What was your mother''s religious preference when you were growing up?  Was it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1058/vshow}{GSS Data Explorer page for \code{marelig}}.
}

\seealso{
Other Religion: 
\code{\link{maden}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}}
}
\concept{Religion}
\keyword{variable}
