% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{marelkid}
\alias{marelkid}
\title{Mothers religion when r was child}
\source{
General Social Survey https://gss.norc.org
}
\description{
marelkid
}
\details{
Question What was your mother's religious preference when you were a child? Was it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \verb{[1]} protestant
\item \verb{[2]} catholic
\item \verb{[3]} jewish
\item \verb{[4]} none
\item \verb{[5]} other
\item \verb{[6]} buddhism
\item \verb{[7]} hinduism
\item \verb{[8]} other eastern religions
\item \verb{[9]} muslim/islam
\item \verb{[10]} orthodox-christian
\item \verb{[11]} native american
\item \verb{[12]} other christian
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4665/vshow}{GSS Data Explorer page for \code{marelkid}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab buddhism \tab hinduism \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 810 \tab 388 \tab 30 \tab 8 \tab 3 \tab 17 \tab 28 \tab 54 \tab 21 \tab 1517 \cr
   1998 \tab 1548 \tab 704 \tab 392 \tab 22 \tab 12 \tab 6 \tab 14 \tab 35 \tab 62 \tab 37 \tab 2832 \cr
   2008 \tab 658 \tab 764 \tab 410 \tab 27 \tab 10 \tab 7 \tab 17 \tab 97 \tab 28 \tab 5 \tab 2023 \cr
   2018 \tab 1173 \tab 600 \tab 379 \tab 20 \tab 7 \tab 11 \tab 16 \tab 102 \tab 25 \tab 15 \tab 2348 \cr
   Total \tab 3537 \tab 2878 \tab 1569 \tab 99 \tab 37 \tab 27 \tab 64 \tab 262 \tab 169 \tab 78 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{relexper}},
\code{\link{religkid}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\keyword{variable}
