% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_479.R
\name{marmakid}
\alias{marmakid}
\title{Single ma can raise kids well as couple}
\source{
General Social Survey https://gss.norc.org
}
\description{
marmakid
}
\details{
Question Do you agree or disagree?
G. A single mother can bring up her child as well as a married couple.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   2022 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4370/vshow}{GSS Data Explorer page for \code{marmakid}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}

Other Marriage: 
\code{\link{femarry}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{memarry}}

Other Parents: 
\code{\link{marpakid}}
}
\concept{ISSP Family & Gender Roles}
\concept{Marriage}
\concept{Parents}
\keyword{variable}
