% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{marno}
\alias{marno}
\title{Close relative marry no. white}
\source{
General Social Survey https://gss.norc.org
}
\description{
marno
}
\details{
Question What about having a close relative marry a Jewish person?  Would you be very in favor of it happening, somewhat In favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
E. A white raised in the North?
(INTERVIEWER: READ OPPOSITE REGION - SOUTH = CENSUS REGIONS - SOUTH ATLANTIC, SOUTH EAST, CENTRAL, SOUTH WEST CENTRAL; NORTH = EVERYWHERE ELSE)
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1225/vshow}{GSS Data Explorer page for \code{marno}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab 914 \tab 51 \tab 77 \tab 268 \tab 35 \tab 16 \tab 6 \tab 5 \tab 1372 \cr
   Total \tab 914 \tab 51 \tab 77 \tab 268 \tab 35 \tab 16 \tab 6 \tab 5 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{lessfare}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marso}},
\code{\link{marwht}},
\code{\link{workfare}}
}
\concept{Intergroup Relations}
\keyword{variable}
