% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_156.R
\name{marnum}
\alias{marnum}
\title{R married more than once?}
\source{
General Social Survey https://gss.norc.org
}
\description{
marnum
}
\details{
Question Have you been married more than once?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1475/vshow}{GSS Data Explorer page for \code{marnum}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1994 \tab 1789 \tab 310 \tab 889 \tab 4 \tab 2992 \cr
   Total \tab 1789 \tab 310 \tab 889 \tab 4 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{alike3}},
\code{\link{alike4}},
\code{\link{alike5}},
\code{\link{alike6}},
\code{\link{alike7}},
\code{\link{alike8}},
\code{\link{fstspdeg}},
\code{\link{fstspedc}},
\code{\link{fstspyr}}
}
\concept{Family Mobility}
\keyword{variable}
