% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{marsame}
\alias{marsame}
\title{Homosexual couples should have the right to marry one another}
\source{
General Social Survey https://gss.norc.org
}
\description{
marsame
}
\details{
Question Do you agree or disagree with the following statement: Homosexual couples should have the right to marry one another.

MARSAME is a rename from MARHOMO. MARSAME is asked in sequence with PILLOK and SPANKING. Hand Card A13 contains punches 1-5.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2021 \tab A/C/- \tab partial \cr
   2022 \tab A/C/- \tab partial \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8056/vshow}{GSS Data Explorer page for \code{marsame}}.
}

\seealso{
Other Core: 
\code{\link{marsame1}}
}
\concept{Core}
\keyword{variable}
