% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{marso}
\alias{marso}
\title{Close relative marry so. white}
\source{
General Social Survey https://gss.norc.org
}
\description{
marso
}
\details{
Question What about having a close relative marry a Jewish person?  Would you be very in favor of it happening, somewhat In favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
F. A white raised in the South?
(INTERVIEWER: READ OPPOSITE REGION - SOUTH = CENSUS REGIONS - SOUTH ATLANTIC, SOUTH EAST, CENTRAL, SOUTH WEST CENTRAL; NORTH = EVERYWHERE ELSE)
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1226/vshow}{GSS Data Explorer page for \code{marso}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab 458 \tab 62 \tab 125 \tab 598 \tab 78 \tab 34 \tab 13 \tab 4 \tab 1372 \cr
   Total \tab 458 \tab 62 \tab 125 \tab 598 \tab 78 \tab 34 \tab 13 \tab 4 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{lessfare}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marwht}},
\code{\link{workfare}}
}
\concept{Intergroup Relations}
\keyword{variable}
