% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{marunion}
\alias{marunion}
\title{Legal marital status}
\source{
General Social Survey https://gss.norc.org
}
\description{
marunion
}
\details{
Question What is your current legal marital status?
}
\section{Values}{

\itemize{
\item \code{1} legally married
\item \code{2} in a civil union or registered domestic partnership
\item \code{3} separated
\item \code{4} not currently married or in a civil union or registered domestic partnership
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3911/vshow}{GSS Data Explorer page for \code{marunion}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab legally married \tab in a civil union or registered domestic partnership \tab separated \tab not currently married or in a civil union or registered domestic partnership \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1879 \tab 32 \tab 8 \tab 6 \tab 65 \tab 2 \tab 31 \tab 2023 \cr
   Total \tab 1879 \tab 32 \tab 8 \tab 6 \tab 65 \tab 2 \tab 31 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{relgendr}},
\code{\link{sexunion}},
\code{\link{toldsxor}}
}
\concept{Sexual Orientation}
\keyword{variable}
