% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{marunion}
\alias{marunion}
\title{Legal marital status}
\source{
General Social Survey https://gss.norc.org
}
\description{
marunion
}
\details{
Question What is your current legal marital status?
}
\section{Values}{

\itemize{
\item \code{1} legally married
\item \code{2} in a civil union or registered domestic partnership
\item \code{3} separated
\item \code{4} not currently married or in a civil union or registered domestic partnership
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3911/vshow}{GSS Data Explorer page for \code{marunion}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{dwell5}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evkid}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{losejob5}},
\code{\link{mykids}},
\code{\link{negjob5}},
\code{\link{othkids}},
\code{\link{relgendr}},
\code{\link{sexsex18}},
\code{\link{sexunion}},
\code{\link{sppart}},
\code{\link{toldsmsx}},
\code{\link{toldsxor}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
