% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{marwht}
\alias{marwht}
\title{R favor close relative marrying white person}
\source{
General Social Survey https://gss.norc.org
}
\description{
marwht
}
\details{
Question What about having a close relative marry a Jewish person?  Would you be very in favor of it happening, somewhat In favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
G. What about having a close relative marry a white person?  Would you be very in favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1227/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab don't know \tab favor \tab neither favor nor oppose \tab no answer \tab oppose \tab strongly favor \tab strongly oppose \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1419 \tab 12 \tab 154 \tab 405 \tab 86 \tab 29 \tab 698 \tab 14 \tab - \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 6 \tab 109 \tab 275 \tab 1 \tab 26 \tab 478 \tab 13 \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 2 \tab 119 \tab 362 \tab 4 \tab 19 \tab 387 \tab 13 \tab - \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 7 \tab 280 \tab 767 \tab 3 \tab 38 \tab 876 \tab 21 \tab - \tab - \tab 4510 \cr
   2008 \tab 694 \tab 4 \tab 191 \tab 548 \tab 2 \tab 25 \tab 544 \tab 15 \tab - \tab - \tab 2023 \cr
   2010 \tab 614 \tab 5 \tab 204 \tab 600 \tab 4 \tab 33 \tab 564 \tab 20 \tab - \tab - \tab 2044 \cr
   2012 \tab 672 \tab 5 \tab 172 \tab 589 \tab 1 \tab 24 \tab 500 \tab 11 \tab - \tab - \tab 1974 \cr
   2014 \tab 863 \tab 8 \tab 238 \tab 765 \tab 4 \tab 40 \tab 607 \tab 13 \tab - \tab - \tab 2538 \cr
   2016 \tab 979 \tab 2 \tab 274 \tab 876 \tab 2 \tab 41 \tab 680 \tab 13 \tab - \tab - \tab 2867 \cr
   2018 \tab 789 \tab 5 \tab 227 \tab 706 \tab 5 \tab 43 \tab 550 \tab 23 \tab - \tab - \tab 2348 \cr
   2021 \tab 1318 \tab - \tab 235 \tab 1709 \tab 2 \tab 54 \tab 666 \tab 30 \tab 18 \tab - \tab 4032 \cr
   2022 \tab 1168 \tab 8 \tab 250 \tab 1319 \tab 4 \tab 47 \tab 724 \tab 23 \tab 1 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 52913 \tab 64 \tab 2453 \tab 8921 \tab 118 \tab 419 \tab 7274 \tab 209 \tab 19 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{lessfare}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{workfare}}

Other Whites: 
\code{\link{livewhts}},
\code{\link{marno}}
}
\concept{Intergroup Relations}
\concept{Whites}
\keyword{variable}
