% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_598.R
\name{masei10educ}
\alias{masei10educ}
\title{Percentage of some college educ in maocc10 based on acs 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
masei10educ
}
\details{
Question Percentage of those who had some college or more education among those who had the same occupation as in MAOCC10 in American Community Survey between 2010 and 2012 (25-64 years old only). This percentage is not from the GSS but it is used to calculate MASEI10.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} i don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5721/vshow}{GSS Data Explorer page for \code{masei10educ}}.
}

\seealso{
Other Occupation: 
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{masei10}},
\code{\link{masei10inc}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{spsei10}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}}
}
\concept{Occupation}
\keyword{variable}
