% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{matesex}
\alias{matesex}
\title{Was 1 of r's partner's spouse or regular}
\source{
General Social Survey https://gss.norc.org
}
\description{
matesex
}
\details{
Question Was one of the partners your husband or wife or regular sexual partner?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5050/vshow}{GSS Data Explorer page for \code{matesex}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{frndsex}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{pikupsex}},
\code{\link{sexfreq}},
\code{\link{sexsex}}

Other Sex: 
\code{\link{acqntsex}},
\code{\link{frndsex}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{pikupsex}},
\code{\link{sexfreq}},
\code{\link{sexsex}}
}
\concept{High Risk Behaviors}
\concept{Sex}
\keyword{variable}
