% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{matime1}
\alias{matime1}
\title{How long does it take to get to where mother live}
\source{
General Social Survey https://gss.norc.org
}
\description{
matime1
}
\details{
Question About how long would it take to get to where your mother lives? Think about the time it usually takes door to door.
}
\section{Values}{

\itemize{
\item \code{1} less than 2 minutes
\item \code{2} less than 15 minutes
\item \code{3} between 15 and 30 minutes
\item \code{4} between 30 minutes and 1 hour
\item \code{5} between 1 and 2 hours
\item \code{6} between 2 and 3 hours
\item \code{7} between 3 and 5 hours
\item \code{8} between 5 and 12 hours
\item \code{9} over 12 hours
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4183/vshow}{GSS Data Explorer page for \code{matime1}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
