% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{matime1}
\alias{matime1}
\title{How long does it take to get to where mother live}
\source{
General Social Survey https://gss.norc.org
}
\description{
matime1
}
\details{
Question About how long would it take to get to where your mother lives? Think about the time it usually takes door to door.
}
\section{Values}{

\itemize{
\item \verb{[1]} less than 2 minutes
\item \verb{[2]} less than 15 minutes
\item \verb{[3]} between 15 and 30 minutes
\item \verb{[4]} between 30 minutes and 1 hour
\item \verb{[5]} between 1 and 2 hours
\item \verb{[6]} between 2 and 3 hours
\item \verb{[7]} between 3 and 5 hours
\item \verb{[8]} between 5 and 12 hours
\item \verb{[9]} over 12 hours
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4183/vshow}{GSS Data Explorer page for \code{matime1}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab less than 2 minutes \tab less than 15 minutes \tab between 15 and 30 minutes \tab between 30 minutes and 1 hour \tab between 1 and 2 hours \tab between 2 and 3 hours \tab between 3 and 5 hours \tab between 5 and 12 hours \tab over 12 hours \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2083 \tab 56 \tab 147 \tab 117 \tab 68 \tab 51 \tab 38 \tab 44 \tab 68 \tab 87 \tab 1 \tab 5 \tab 2765 \cr
   Total \tab 2083 \tab 56 \tab 147 \tab 117 \tab 68 \tab 51 \tab 38 \tab 44 \tab 68 \tab 87 \tab 1 \tab 5 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
