% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{mawk16}
\alias{mawk16}
\title{Mothers employment when r growing up}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawk16
}
\details{
Question Did your mother ever work for pay for as long as a year, after she was married?
C. Did she work for as long as a year around the time you were 16?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab - \tab full \cr
   1974 \tab - \tab full \cr
   1975 \tab - \tab full \cr
   1976 \tab - \tab full \cr
   1977 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/91/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 913 \tab 6 \tab 185 \tab 1 \tab 399 \tab - \tab 1504 \cr
   1974 \tab 882 \tab 2 \tab 186 \tab 2 \tab 412 \tab - \tab 1484 \cr
   1975 \tab 794 \tab 4 \tab 226 \tab 19 \tab 447 \tab - \tab 1490 \cr
   1976 \tab 802 \tab 5 \tab 233 \tab 21 \tab 438 \tab - \tab 1499 \cr
   1977 \tab 877 \tab 2 \tab 201 \tab 31 \tab 419 \tab - \tab 1530 \cr
   1978 \tab 776 \tab 2 \tab 253 \tab 10 \tab 491 \tab - \tab 1532 \cr
   1980 \tab 743 \tab 2 \tab 259 \tab 17 \tab 447 \tab - \tab 1468 \cr
   1982 \tab 944 \tab 2 \tab 216 \tab 67 \tab 631 \tab - \tab 1860 \cr
   1983 \tab 731 \tab 4 \tab 273 \tab 7 \tab 584 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 62342 \tab 29 \tab 2032 \tab 175 \tab 4268 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{born}},
\code{\link{famdif16}},
\code{\link{family16}},
\code{\link{incom16}},
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}},
\code{\link{parborn}}

Other Labor Force Status: 
\code{\link{mawkbaby}},
\code{\link{mawkborn}},
\code{\link{mawork}},
\code{\link{mawrkgrw}}
}
\concept{Core}
\concept{Labor Force Status}
\keyword{variable}
