% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{mawork14}
\alias{mawork14}
\title{Did mom work before r was 14 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawork14
}
\details{
Question Did your mother ever work for pay for as long as one year after you were born and before you were 14?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} didnt live with mom
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4392/vshow}{GSS Data Explorer page for \code{mawork14}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabit}},
\code{\link{divhubby}},
\code{\link{earnsmor}},
\code{\link{evdiv}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\concept{Labor Force Status}
\keyword{variable}
