% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_612.R
\name{mawrkslffam}
\alias{mawrkslffam}
\title{Mother (or youngest same-sex parent) owned family business or farm}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawrkslffam
}
\details{
Question At this job, did your mother work in her own family business or family farm?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7431/vshow}{GSS Data Explorer page for \code{mawrkslffam}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 1169 \tab 245 \tab 2613 \tab 1 \tab 4 \tab 4032 \cr
   2022 \tab 966 \tab 240 \tab 2336 \tab - \tab 2 \tab 3544 \cr
   Total \tab 2135 \tab 485 \tab 4949 \tab 1 \tab 6 \tab 7576 \cr
}
}

\keyword{variable}
