% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_477.R
\name{mawrkwrm}
\alias{mawrkwrm}
\title{Working mom can have a warm rel w kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawrkwrm
}
\details{
Question Do you agree or disagree ...
A. A working mother can establish just as warm and secure a relationship with her children as a mother who does not work.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2012 \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4343/vshow}{GSS Data Explorer page for \code{mawrkwrm}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1988 \tab 67 \tab 394 \tab 513 \tab 121 \tab 253 \tab 102 \tab 11 \tab 20 \tab 1481 \cr
   1994 \tab 1545 \tab 418 \tab 592 \tab 72 \tab 273 \tab 70 \tab 9 \tab 13 \tab 2992 \cr
   2002 \tab 1594 \tab 468 \tab 383 \tab 124 \tab - \tab 184 \tab 11 \tab 1 \tab 2765 \cr
   2012 \tab 672 \tab 491 \tab 523 \tab 51 \tab 171 \tab 47 \tab 14 \tab 5 \tab 1974 \cr
   Total \tab 3878 \tab 1771 \tab 2011 \tab 368 \tab 697 \tab 403 \tab 45 \tab 39 \tab 9212 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{famsuffr}},
\code{\link{fejobind}},
\code{\link{hapifwrk}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywrk}},
\code{\link{kidsuffr}},
\code{\link{twoincs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
