% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{mayrdied}
\alias{mayrdied}
\title{Yr r mother died}
\source{
General Social Survey https://gss.norc.org
}
\description{
mayrdied
}
\details{
Question In what year did she die?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1456/vshow}{GSS Data Explorer page for \code{mayrdied}}.
}

\seealso{
Other Family Mobility: 
\code{\link{malive1}},
\code{\link{mayrborn}},
\code{\link{palive1}},
\code{\link{payrborn}}
}
\concept{Family Mobility}
\keyword{variable}
