% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_226.R
\name{mcsds5}
\alias{mcsds5}
\title{I have never deliberately said something that hurt someones feelings}
\source{
General Social Survey https://gss.norc.org
}
\description{
mcsds5
}
\details{
Question Next are a number of statements concerning personal attitudes and traits. Please tell me whether the statement is true or false as it pertains to you personally: I have never deliberately said something that hurt someone~s feelings.
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7153/vshow}{GSS Data Explorer page for \code{mcsds5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 458 \tab 684 \tab 4 \tab 27 \tab 2348 \cr
   Total \tab 1175 \tab 458 \tab 684 \tab 4 \tab 27 \tab 2348 \cr
}
}

\keyword{variable}
