% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_633.R
\name{mditate1}
\alias{mditate1}
\title{How often do you meditate?}
\source{
General Social Survey https://gss.norc.org
}
\description{
mditate1
}
\details{
Question How often do you meditate?
}
\section{Values}{

\itemize{
\item \code{1} at least once a day
\item \code{2} almost every day
\item \code{3} once or twice a week
\item \code{4} once or twice a month
\item \code{5} a few times per year
\item \code{6} once a year or less
\item \code{7} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7537/vshow}{GSS Data Explorer page for \code{mditate1}}.
}

\keyword{variable}
