% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{medbest}
\alias{medbest}
\title{How likely to get best treatment avail in us}
\source{
General Social Survey https://gss.norc.org
}
\description{
medbest
}
\details{
Question How likely is it that if you become seriously ill, you would get or not get the best treatment available in the United States?
}
\section{Values}{

\itemize{
\item \verb{[1]} it's certain r would get
\item \verb{[2]} it's likely r would get
\item \verb{[3]} equal chance of getting or not getting
\item \verb{[4]} it's likely r would not get
\item \verb{[5]} it's certain r would not get
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7830/vshow}{GSS Data Explorer page for \code{medbest}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab it's certain r would get \tab it's likely r would get \tab equal chance of getting or not getting \tab it's likely r would not get \tab it's certain r would not get \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2382 \tab 242 \tab 418 \tab 284 \tab 130 \tab 33 \tab 48 \tab 3 \tab 4 \tab 3544 \cr
   Total \tab 2382 \tab 242 \tab 418 \tab 284 \tab 130 \tab 33 \tab 48 \tab 3 \tab 4 \tab 3544 \cr
}
}

\keyword{variable}
