% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{medcare1}
\alias{medcare1}
\title{Who r thinks should be most responsible for cost of xs medical care}
\source{
General Social Survey https://gss.norc.org
}
\description{
medcare1
}
\details{
Question In your opinion, who should be most responsible for paying the cost of NAME's medical care, including mental health care and treatment?
}
\section{Values}{

\itemize{
\item \code{1} name, him-herself
\item \code{2} his-her family
\item \code{3} government
\item \code{4} insurance
\item \code{5} private charity
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2036/vshow}{GSS Data Explorer page for \code{medcare1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab name, him-herself \tab his-her family \tab government \tab insurance \tab private charity \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 550 \tab 105 \tab 147 \tab 515 \tab 35 \tab 74 \tab 18 \tab 2904 \cr
   2006 \tab 3073 \tab 400 \tab 93 \tab 191 \tab 675 \tab 23 \tab 38 \tab 17 \tab 4510 \cr
   Total \tab 4533 \tab 950 \tab 198 \tab 338 \tab 1190 \tab 58 \tab 112 \tab 35 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare2}},
\code{\link{mntlas1}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}},
\code{\link{spmentl}}
}
\concept{Mental Health}
\keyword{variable}
