% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_580.R
\name{mediarel}
\alias{mediarel}
\title{Watch a religious program on television-radio?}
\source{
General Social Survey https://gss.norc.org
}
\description{
mediarel
}
\details{
Question During the last seven days did you do the following:
B. Watch a religious program on television or listen to a religious program on the radio?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5319/vshow}{GSS Data Explorer page for \code{mediarel}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1953 \tab 213 \tab 668 \tab 1 \tab 69 \tab 2904 \cr
   1998 \tab 1387 \tab 400 \tab 1006 \tab 1 \tab 38 \tab 2832 \cr
   Total \tab 3340 \tab 613 \tab 1674 \tab 2 \tab 107 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attreg}},
\code{\link{friday}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
