% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_227.R
\name{meditate}
\alias{meditate}
\title{How often meditate}
\source{
General Social Survey https://gss.norc.org
}
\description{
meditate
}
\details{
Question Within your religious or spiritual tradition, how often do you meditate?
}
\section{Values}{

\itemize{
\item \verb{[1]} mre thn once/day
\item \verb{[2]} once a day
\item \verb{[3]} a few times/week
\item \verb{[4]} once a week
\item \verb{[5]} few times/month
\item \verb{[6]} once a month
\item \verb{[7]} less once/mnth
\item \verb{[8]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2091/vshow}{GSS Data Explorer page for \code{meditate}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab mre thn once/day \tab once a day \tab a few times/week \tab once a week \tab few times/month \tab once a month \tab less once/mnth \tab never \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 123 \tab 188 \tab 130 \tab 68 \tab 60 \tab 46 \tab 95 \tab 660 \tab 47 \tab 28 \tab 2832 \cr
   Total \tab 1387 \tab 123 \tab 188 \tab 130 \tab 68 \tab 60 \tab 46 \tab 95 \tab 660 \tab 47 \tab 28 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{privpray}}
}
\concept{Religion and Culture}
\keyword{variable}
