% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{medpay}
\alias{medpay}
\title{Couldn't get care because couldn't pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
medpay
}
\details{
Question During the past 12 months did it ever happen that you did not get the medical treatment you needed because... You could not pay for it?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} did not need medical treatment
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7827/vshow}{GSS Data Explorer page for \code{medpay}}.
}

\keyword{variable}
