% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_390.R
\name{medtreat}
\alias{medtreat}
\title{How scientific is medical treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
medtreat
}
\details{
Question E. Medical treatment. (Is medical treatment...?)
}
\section{Values}{

\itemize{
\item \verb{[1]} very scientific
\item \verb{[2]} pretty scientific
\item \verb{[3]} not too scientific
\item \verb{[4]} not scientific at all
\item \verb{[5]} haven't heard of it (volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3541/vshow}{GSS Data Explorer page for \code{medtreat}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very scientific \tab pretty scientific \tab not too scientific \tab not scientific at all \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1469 \tab 382 \tab 93 \tab 14 \tab 3 \tab 9 \tab 4 \tab 1974 \cr
   Total \tab 1469 \tab 382 \tab 93 \tab 14 \tab 3 \tab 9 \tab 4 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{architct}},
\code{\link{cmprgmng}},
\code{\link{engnring}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{lawenfrc}},
\code{\link{marrcoun}},
\code{\link{slsmnshp}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
