% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{medwtlst}
\alias{medwtlst}
\title{Couldn't get care because waitlist too long}
\source{
General Social Survey https://gss.norc.org
}
\description{
medwtlst
}
\details{
Question (During the past 12 months did it ever happen that you did not get the medical treatment you needed because...) The waiting list was too long?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} did not need medical treatment
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7829/vshow}{GSS Data Explorer page for \code{medwtlst}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab did not need medical treatment \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2381 \tab 147 \tab 836 \tab 173 \tab 1 \tab 2 \tab 4 \tab 3544 \cr
   Total \tab 2381 \tab 147 \tab 836 \tab 173 \tab 1 \tab 2 \tab 4 \tab 3544 \cr
}
}

\keyword{variable}
