% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{meetf2f1}
\alias{meetf2f1}
\title{How often meet face to face with customers, clients, patients}
\source{
General Social Survey https://gss.norc.org
}
\description{
meetf2f1
}
\details{
Question How often do you meet face to face with customers, clients, patients, or others to whom you provide products or services?  Would you say….
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} once per year or less, or
\item \verb{[3]} several times per year
\item \verb{[4]} several times per month
\item \verb{[5]} several times per week
\item \verb{[6]} daily
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3890/vshow}{GSS Data Explorer page for \code{meetf2f1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab never \tab once per year or less, or \tab several times per year \tab several times per month \tab several times per week \tab daily \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 139 \tab 28 \tab 44 \tab 69 \tab 75 \tab 485 \tab 5 \tab 13 \tab 2023 \cr
   Total \tab 1165 \tab 139 \tab 28 \tab 44 \tab 69 \tab 75 \tab 485 \tab 5 \tab 13 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
