% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{meetings}
\alias{meetings}
\title{How many people seen at church or group meeting}
\source{
General Social Survey https://gss.norc.org
}
\description{
meetings
}
\details{
Question Of these____friends and relatives, about how many do you stay in contact with by:
D. Seeing them at meetings or events related to church, clubs, or other groups?
}
\section{Values}{

\itemize{
\item \code{1} 0
\item \code{2} 1 - 2
\item \code{3} 3 - 5
\item \code{4} 6 - 10
\item \code{5} 11 - 15
\item \code{6} 16-25
\item \code{7} 26-50
\item \code{8} 50+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2547/vshow}{GSS Data Explorer page for \code{meetings}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 - 2 \tab 3 - 5 \tab 6 - 10 \tab 11 - 15 \tab 16-25 \tab 26-50 \tab 50+ \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1884 \tab 364 \tab 135 \tab 122 \tab 103 \tab 51 \tab 62 \tab 27 \tab 39 \tab 2 \tab 28 \tab 2817 \cr
   2002 \tab 74 \tab 982 \tab 292 \tab 384 \tab 369 \tab 174 \tab 215 \tab 153 \tab 105 \tab 4 \tab 13 \tab 2765 \cr
   2004 \tab 1873 \tab 428 \tab 100 \tab 127 \tab 110 \tab 46 \tab 55 \tab 31 \tab 28 \tab 2 \tab 12 \tab 2812 \cr
   Total \tab 3831 \tab 1774 \tab 527 \tab 633 \tab 582 \tab 271 \tab 332 \tab 211 \tab 172 \tab 8 \tab 53 \tab 8394 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}},
\code{\link{polinfgn}}
}
\concept{Information Society}
\keyword{variable}
