% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{mehhwork}
\alias{mehhwork}
\title{Men ought to do more hh work than they do now}
\source{
General Social Survey https://gss.norc.org
}
\description{
mehhwork
}
\details{
Question To what extent do you agree or disagree...?
A. Men ought to do a larger share of household work than they do now.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4427/vshow}{GSS Data Explorer page for \code{mehhwork}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1594 \tab 250 \tab 482 \tab 311 \tab 105 \tab 22 \tab 1 \tab 2765 \cr
   Total \tab 1594 \tab 250 \tab 482 \tab 311 \tab 105 \tab 22 \tab 1 \tab 2765 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mekdcare}},
\code{\link{repairs1}},
\code{\link{shopfood}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
