% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_171.R
\name{meltpot}
\alias{meltpot}
\title{R attitude toward mixing cultures in us}
\source{
General Social Survey https://gss.norc.org
}
\description{
meltpot
}
\details{
Question Some people say that it is better for America if different racial and ethnic groups maintain their distinct cultures. Others say that it is better if groups change so that they blend into the larger society as in the idea of a melting pot. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that racial and ethnic groups should maintain their distinct cultures, and a score of 7 as meaning that groups should change so that they blend into the larger society. What score between 1 and 7 comes closest to the way you feel? (CIRCLE ONE):
}
\section{Values}{

\itemize{
\item \code{1} maintain distinct cultures
\item \code{7} blend in to larger society
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1599/vshow}{GSS Data Explorer page for \code{meltpot}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab maintain distinct cultures \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab blend in to larger society \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 191 \tab 102 \tab 166 \tab 424 \tab 230 \tab 114 \tab 208 \tab 33 \tab 6 \tab 2992 \cr
   2000 \tab 1419 \tab 120 \tab 132 \tab 173 \tab 438 \tab 185 \tab 140 \tab 137 \tab 62 \tab 11 \tab 2817 \cr
   Total \tab 2937 \tab 311 \tab 234 \tab 339 \tab 862 \tab 415 \tab 254 \tab 345 \tab 95 \tab 17 \tab 5809 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{amissue}},
\code{\link{amissue1}},
\code{\link{amproud}},
\code{\link{amrank}},
\code{\link{ethid1}},
\code{\link{ethissu1}},
\code{\link{ethissue}},
\code{\link{gvtapart}},
\code{\link{gvtmelt}}
}
\concept{Multiculturalism}
\keyword{variable}
