% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{meltpot1}
\alias{meltpot1}
\title{Better to maintain distinct cultures}
\source{
General Social Survey https://gss.norc.org
}
\description{
meltpot1
}
\details{
Question Some people say that it is better for a country if different racial and ethnic groups maintain their distinct customs and traditions. Others say that it is better if these groups adapt and blend into the larger society. Which of these views comes closer to your own?
}
\section{Values}{

\itemize{
\item \code{1} better for society maintain distinct
\item \code{2} better if groups adapt and blend into
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4858/vshow}{GSS Data Explorer page for \code{meltpot1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab better for society maintain distinct \tab better if groups adapt and blend into \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 399 \tab 557 \tab 348 \tab 63 \tab 2904 \cr
   2004 \tab 1596 \tab 554 \tab 615 \tab 28 \tab 19 \tab 2812 \cr
   2014 \tab 1264 \tab 478 \tab 597 \tab 191 \tab 8 \tab 2538 \cr
   Total \tab 4397 \tab 1431 \tab 1769 \tab 567 \tab 90 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{refugees}},
\code{\link{res161}}
}
\concept{ISSP National Identity}
\keyword{variable}
