% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memfarm}
\alias{memfarm}
\title{Membership in farm organization}
\source{
General Social Survey https://gss.norc.org
}
\description{
memfarm
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
L. Farm organizations
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab B/C/- \tab partial \cr
   1989 \tab B/C/- \tab partial \cr
   1990 \tab B/C/- \tab partial \cr
   1991 \tab B/C/- \tab partial \cr
   1993 \tab B/C/- \tab partial \cr
   1994 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/974/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1 \tab 1399 \tab 21 \tab 63 \tab - \tab 1484 \cr
   1975 \tab - \tab 2 \tab 1398 \tab 29 \tab 61 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1 \tab 1455 \tab 14 \tab 60 \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab 1460 \tab 18 \tab 54 \tab - \tab 1532 \cr
   1980 \tab - \tab 1 \tab 1379 \tab 29 \tab 59 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab 1519 \tab 9 \tab 71 \tab - \tab 1599 \cr
   1984 \tab - \tab 2 \tab 1387 \tab 22 \tab 62 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab 1403 \tab 11 \tab 56 \tab - \tab 1470 \cr
   1987 \tab - \tab 1 \tab 1742 \tab 21 \tab 55 \tab - \tab 1819 \cr
   1988 \tab 484 \tab - \tab 950 \tab 13 \tab 34 \tab - \tab 1481 \cr
   1989 \tab 502 \tab - \tab 970 \tab 33 \tab 32 \tab - \tab 1537 \cr
   1990 \tab 473 \tab - \tab 847 \tab 9 \tab 43 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 1 \tab 971 \tab 12 \tab 33 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 1 \tab 1005 \tab 6 \tab 45 \tab - \tab 1606 \cr
   1994 \tab 2481 \tab - \tab 490 \tab 2 \tab 19 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1340 \tab - \tab 1423 \tab 5 \tab 44 \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 47993 \tab 10 \tab 19798 \tab 254 \tab 791 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{memchurh}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{memprof}},
\code{\link{memschl}}

Other Voluntary Associations: 
\code{\link{memchurh}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{memprof}},
\code{\link{memschl}}
}
\concept{Agriculture}
\concept{Socio-political participation}
\concept{Voluntary Associations}
\keyword{variable}
